Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  0  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  3  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  1  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 12.00581                                                 fv = 20.64377                                                 fv = 60.48594                                                 fv = 163.6402                                                 fv = 12028.91                                                 fv = 18.2474                                                 fv = 5110065                                                 fv = 331580.5                                                 fv = 18.30568                                                 fv = 240.4078                                                 fv = 1.340781e+154                                                 fv = 11.61112                                                 fv = 15.83537                                                 fv = 98.88284                                                 fv = 1.340781e+154                                                 fv = 13.08282                                                 fv = 11.15957                                                 fv = 8.556558                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.67893                                                 fv = 9.788179                                                 fv = 10.50757                                                 fv = 1.340781e+154                                                 fv = 10.39095                                                 fv = 1.340781e+154                                                 fv = 22.05843                                                 fv = 36.6807                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.698612                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.793119                                                 fv = 12.6474                                                 fv = 1.340781e+154                                                 fv = 8.556558                                                 fv = 14.92742                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.354773                                                 fv = 1.340781e+154                                                 fv = 10.39008                                                 fv = 138.9964                                                 fv = 1618.378                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.061686                                                 fv = 1.340781e+154                                                 fv = 6.061686                                                 fv = 13.80944                                                 fv = 7.435668                                                 fv = 4.698612                                                 fv = 4.972938                                                 fv = 6.1721                                                 fv = 5.314878                                                 fv = 4.793557                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.698612                                                 fv = 1.340781e+154                                                 fv = 4.698612                                                 fv = 1.340781e+154                                                 fv = 8.688112                                                 fv = 8.556558                                                 fv = 1560.594                                                 fv = 8.75951 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 92.91918                                                 fv = 4.314177                                                 fv = 5.53345                                                 fv = 11.02196                                                 fv = 1.340781e+154                                                 fv = 8.599384                                                 fv = 4.949851                                                 fv = 5.395238                                                 fv = 4.98613                                                 fv = 5.395977                                                 fv = 41.62583                                                 fv = 4.314177                                                 fv = 4.314177                                                 fv = 4.314177                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.870003                                                 fv = 5.819921                                                 fv = 1.340781e+154                                                 fv = 8.31268                                                 fv = 3.733478                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 53502570196                                                 fv = 3.627307                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.23749                                                 fv = 1.340781e+154                                                 fv = 4.804275                                                 fv = 4.055798                                                 fv = 4.218048                                                 fv = 3.627307                                                 fv = 4.555968                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.728842                                                 fv = 3.627307                                                 fv = 7.688184                                                 fv = 3.627307                                                 fv = 41720640499                                                 fv = 3.944897                                                 fv = 4.076503                                                 fv = 13.71824                                                 fv = 24.59955                                                 fv = 1.340781e+154                                                 fv = 4.073512                                                 fv = 4.962129                                                 fv = 3.627307                                                 fv = 4.228289                                                 fv = 3.627307                                                 fv = 4.30189                                                 fv = 4.555425                                                 fv = 1.340781e+154                                                 fv = 4.108981                                                 fv = 4.029569                                                 fv = 5.259193                                                 fv = 5.677135 run # 0: f=3.627307392
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 124.8133                                                 fv = 46.8464                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 33.23885                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.0982                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 207.0708                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9346.183                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.082658                                                 fv = 774490.9                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 71.93737                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.00642                                                 fv = 10.75839                                                 fv = 15.00642                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.10143                                                 fv = 1.340781e+154                                                 fv = 29599.02                                                 fv = 1.340781e+154                                                 fv = 4.697592                                                 fv = 15.87467                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 25830.55                                                 fv = 10.51433                                                 fv = 6.665389                                                 fv = 1.340781e+154                                                 fv = 4.697592                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 22.96725                                                 fv = 8.660251                                                 fv = 7.008816                                                 fv = 18.4444                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 32.08726                                                 fv = 1.340781e+154                                                 fv = 465.8103                                                 fv = 492.4792                                                 fv = 8.526121                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.107873                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.970398                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.294505                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.21589                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 123.2539                                                 fv = 1.340781e+154                                                 fv = 5.396253                                                 fv = 4.697592                                                 fv = 1.340781e+154                                                 fv = 1329.222                                                 fv = 882.2466                                                 fv = 10.63672                                                 fv = 11.44651                                                 fv = 6.262377                                                 fv = 1.340781e+154                                                 fv = 12.23895                                                 fv = 1.340781e+154                                                 fv = 4.697592                                                 fv = 9.279711                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.969746                                                 fv = 147.1167                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.268202                                                 fv = 43.96831                                                 fv = 11.16214                                                 fv = 8.751654                                                 fv = 7.571498                                                 fv = 4.697592                                                 fv = 8.600098                                                 fv = 1.340781e+154                                                 fv = 7.294505                                                 fv = 8.747053                                                 fv = 10.75222                                                 fv = 1.340781e+154                                                 fv = 7.294505                                                 fv = 42.85635                                                 fv = 1.340781e+154                                                 fv = 17.01641                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.717372                                                 fv = 6.268202                                                 fv = 4.179485                                                 fv = 116.3652                                                 fv = 13.77139                                                 fv = 13.83558                                                 fv = 4.179485                                                 fv = 3.897979                                                 fv = 15.57198                                                 fv = 1.340781e+154                                                 fv = 6.431376                                                 fv = 4.380866                                                 fv = 4.380866                                                 fv = 3.897979                                                 fv = 3.897979                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.126399                                                 fv = 3.897979                                                 fv = 3.897979                                                 fv = 5.450777                                                 fv = 8.632093                                                 fv = 6.011263                                                 fv = 3.897979                                                 fv = 5.655678                                                 fv = 8.96332                                                 fv = 5.689231                                                 fv = 4.946365                                                 fv = 10.06655                                                 fv = 1.340781e+154 run # 1: f=3.897978992
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 2: f=1.340780793e+154
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3500.874                                                 fv = 46.8464                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 46.8464                                                 fv = 46.8464                                                 fv = 111489.2                                                 fv = 14.49442                                                 fv = 786149.7                                                 fv = 41.05929                                                 fv = 1.340781e+154                                                 fv = 14.39913                                                 fv = 63239457                                                 fv = 813.4612                                                 fv = 1.340781e+154                                                 fv = 14.39913                                                 fv = 597.7507                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4803085564                                                 fv = 1.340781e+154                                                 fv = 21.23152                                                 fv = 21772.7                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1369.474                                                 fv = 1.340781e+154                                                 fv = 4936.846                                                 fv = 4936.846                                                 fv = 21.33502                                                 fv = 38662088475                                                 fv = 14.39913                                                 fv = 65.8472                                                 fv = 1.340781e+154                                                 fv = 14.03045                                                 fv = 1.340781e+154                                                 fv = 13.93361                                                 fv = 13.93361                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.01812                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.24224                                                 fv = 1.340781e+154                                                 fv = 2.35234e+12                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.9244                                                 fv = 140.3588                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 13.93361                                                 fv = 15.7827                                                 fv = 1847547607                                                 fv = 1.340781e+154                                                 fv = 987730528                                                 fv = 19.33303                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 111.9702                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 21866.4                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 21.4307                                                 fv = 1.340781e+154                                                 fv = 13.93361                                                 fv = 11.48186                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 19.41581                                                 fv = 14350.64                                                 fv = 14937.89                                                 fv = 10.79123                                                 fv = 8700.145                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 30.10066                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 20445.23                                                 fv = 11.48186                                                 fv = 1.340781e+154                                                 fv = 21.08565                                                 fv = 30.10066                                                 fv = 30.10066                                                 fv = 10.79123                                                 fv = 615.2294                                                 fv = 10.79123                                                 fv = 1.340781e+154                                                 fv = 13.56005                                                 fv = 1108.468                                                 fv = 398.2444                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 137032185958                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1060.785                                                 fv = 10.79123                                                 fv = 1060.785                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 43.74078                                                 fv = 1.340781e+154                                                 fv = 82.40021                                                 fv = 11.10989                                                 fv = 10.79123                                                 fv = 15.52454                                                 fv = 56.58982                                                 fv = 11.48186                                                 fv = 65.42858                                                 fv = 10.79123                                                 fv = 1.340781e+154                                                 fv = 56.58982                                                 fv = 1.340781e+154                                                 fv = 64.33838                                                 fv = 18324175                                                 fv = 1.340781e+154                                                 fv = 16.34963                                                 fv = 11.48186                                                 fv = 22.34051                                                 fv = 11.48186                                                 fv = 14.86739                                                 fv = 38.09847                                                 fv = 11.203                                                 fv = 18.86709                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 77.15683                                                 fv = 10.79123                                                 fv = 1108.468                                                 fv = 7989086                                                 fv = 1108.468                                                 fv = 10.79123                                                 fv = 1.340781e+154                                                 fv = 725.8728                                                 fv = 15.5426                                                 fv = 161126692678                                                 fv = 18.35408                                                 fv = 1.340781e+154                                                 fv = 11.81859                                                 fv = 564822550332                                                 fv = 21.92272                                                 fv = 11.81859                                                 fv = 12.97404                                                 fv = 10.79123                                                 fv = 1.340781e+154                                                 fv = 21.77893                                                 fv = 10.79123                                                 fv = 10.79123                                                 fv = 10.79123                                                 fv = 581.8177 run # 3: f=10.79122594
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 209690.9                                                 fv = 5025.673                                                 fv = 14.03045                                                 fv = 1.340781e+154                                                 fv = 8.64215                                                 fv = 45.26204                                                 fv = 1.340781e+154                                                 fv = 5.027518                                                 fv = 4663.625                                                 fv = 1.340781e+154                                                 fv = 8700162                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.905192                                                 fv = 4.941353                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 438.3652                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.941353                                                 fv = 1.340781e+154                                                 fv = 4.941353                                                 fv = 4.941353                                                 fv = 1.340781e+154                                                 fv = 29.51225                                                 fv = 35.61191                                                 fv = 1.340781e+154                                                 fv = 198.8346                                                 fv = 15.05172                                                 fv = 6.780838                                                 fv = 3.071152e+13                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.941353                                                 fv = 1.340781e+154                                                 fv = 9.143428                                                 fv = 8834370217                                                 fv = 1484.485                                                 fv = 1.340781e+154                                                 fv = 69.57192                                                 fv = 5.326721                                                 fv = 1.340781e+154                                                 fv = 29.58228                                                 fv = 7.883449                                                 fv = 128.9783                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 708.6317                                                 fv = 303860.9                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.856134                                                 fv = 1.340781e+154                                                 fv = 8.573164                                                 fv = 257.9423                                                 fv = 4.941353                                                 fv = 4.941353                                                 fv = 4.941353                                                 fv = 4.941353                                                 fv = 7.770414                                                 fv = 8.794611                                                 fv = 7.883449                                                 fv = 11.16545                                                 fv = 5.248603                                                 fv = 5.1131                                                 fv = 24.39615                                                 fv = 22.51053                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.366435                                                 fv = 1.340781e+154                                                 fv = 5.611292                                                 fv = 12.62665                                                 fv = 6.366435                                                 fv = 4.941353                                                 fv = 4.941353                                                 fv = 4.941353                                                 fv = 8.22065 run # 4: f=4.941352962
#> 
#> bb eval : 746
#> best    : 3.627307392
#> worst   : 1.340780793e+154
#> solution: x = ( 1 1 1 0 1 2 0 0 0 0 1 1 1 1 1 1 1 1 1 1 ) f(x) = 3.627307392
#> 
#>                                                 fv = 3.627307                                                                       
#> Warning:  optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (4): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max
#> Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...          

# Score the predictions
predictions$score()
#> regr.mse 
#> 41.20969