Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  0  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  2  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  1  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 25.142                                                 fv = 14.70932                                                 fv = 50075.14                                                 fv = 30.01516                                                 fv = 1.340781e+154                                                 fv = 16.85075                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 56.14934                                                 fv = 14.70932                                                 fv = 1.340781e+154                                                 fv = 1503.415                                                 fv = 16.85075                                                 fv = 1.340781e+154                                                 fv = 18.61173                                                 fv = 18.61173                                                 fv = 118.8104                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.795593                                                 fv = 13.56913                                                 fv = 1.340781e+154                                                 fv = 7.459703                                                 fv = 20.99499                                                 fv = 1.340781e+154                                                 fv = 695344373407                                                 fv = 7.459703                                                 fv = 121904287557                                                 fv = 1.340781e+154                                                 fv = 9.767706                                                 fv = 1.340781e+154                                                 fv = 8.281964                                                 fv = 1.340781e+154                                                 fv = 6.680043                                                 fv = 89.53803                                                 fv = 1.340781e+154                                                 fv = 56.02459                                                 fv = 20.25535                                                 fv = 1.340781e+154                                                 fv = 7.686929                                                 fv = 28.44313                                                 fv = 6.680043                                                 fv = 1.340781e+154                                                 fv = 7.049148                                                 fv = 8.186645                                                 fv = 1.340781e+154                                                 fv = 67.83485                                                 fv = 1.340781e+154                                                 fv = 6.680043                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 22.13676                                                 fv = 8.237935                                                 fv = 1.340781e+154                                                 fv = 5704234                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.223102                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.164605                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.286868                                                 fv = 2.054161e+13                                                 fv = 6.170221                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.814294                                                 fv = 6.809529                                                 fv = 1.340781e+154                                                 fv = 6.680043                                                 fv = 55.79103                                                 fv = 6678166234                                                 fv = 11.21753                                                 fv = 6.164605                                                 fv = 14.17965                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.18097                                                 fv = 1.340781e+154                                                 fv = 12718.86                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.741993                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.677969                                                 fv = 6.164605                                                 fv = 10.55075                                                 fv = 9.923445                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.951235                                                 fv = 6.391932                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.800312                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.686929                                                 fv = 1.340781e+154                                                 fv = 84.25138                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 1.340781e+154                                                 fv = 15260.53                                                 fv = 16.78402                                                 fv = 5.800312                                                 fv = 5.800312                                                 fv = 8.186645                                                 fv = 1.340781e+154                                                 fv = 8.186645                                                 fv = 1.340781e+154                                                 fv = 1611.405                                                 fv = 1.340781e+154                                                 fv = 307.624                                                 fv = 1623.193                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.0387                                                 fv = 8.827474                                                 fv = 1.340781e+154                                                 fv = 5.800312                                                 fv = 81.65206                                                 fv = 15.71252                                                 fv = 1.340781e+154                                                 fv = 31.45476                                                 fv = 7.998798                                                 fv = 1.340781e+154                                                 fv = 67.83485                                                 fv = 7.050669                                                 fv = 1.340781e+154                                                 fv = 28.43929                                                 fv = 1.340781e+154                                                 fv = 5.800312                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.066602                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.926703                                                 fv = 41.04225                                                 fv = 8.751124                                                 fv = 8.309193                                                 fv = 7.024721                                                 fv = 6.376403                                                 fv = 6.593373                                                 fv = 11.34684                                                 fv = 484600.2                                                 fv = 5.800312                                                 fv = 1.340781e+154                                                 fv = 5.800312                                                 fv = 1.340781e+154                                                 fv = 8.615518                                                 fv = 5.800312                                                 fv = 9.339935                                                 fv = 10.0104                                                 fv = 29.31743                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 159916.6                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.34098                                                 fv = 7.964651                                                 fv = 1.340781e+154                                                 fv = 26.54224                                                 fv = 7.991785                                                 fv = 5.800312                                                 fv = 5.800312                                                 fv = 36.0763                                                 fv = 78.28583                                                 fv = 5.800312                                                 fv = 5.800312                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.593373                                                 fv = 11.40848                                                 fv = 12.57364                                                 fv = 12.55493                                                 fv = 6.717744                                                 fv = 10.64681                                                 fv = 1.340781e+154                                                 fv = 7.341563                                                 fv = 5.577938                                                 fv = 7.437689                                                 fv = 1.340781e+154                                                 fv = 6.82957                                                 fv = 5.577938                                                 fv = 5.922399                                                 fv = 5.471365                                                 fv = 1.340781e+154                                                 fv = 7.341563                                                 fv = 5.471365                                                 fv = 5.471365                                                 fv = 5.922399                                                 fv = 5.471365                                                 fv = 5.924379                                                 fv = 1.340781e+154                                                 fv = 5.471365                                                 fv = 5.289342                                                 fv = 1.340781e+154                                                 fv = 8.058366                                                 fv = 5.289342                                                 fv = 5.289342                                                 fv = 5.289342                                                 fv = 5.289342                                                 fv = 5.289342                                                 fv = 5.289342                                                 fv = 6.120081                                                 fv = 7.080046                                                 fv = 6.9007                                                 fv = 5.12684                                                 fv = 1.340781e+154                                                 fv = 5.12684                                                 fv = 5.12684                                                 fv = 5.12684                                                 fv = 5.12684                                                 fv = 5.12684                                                 fv = 5.12684                                                 fv = 5.725141                                                 fv = 5.525145                                                 fv = 5.522092                                                 fv = 7.276111                                                 fv = 5.550716                                                 fv = 9.630197                                                 fv = 7.127688                                                 fv = 1.340781e+154                                                 fv = 6.827191                                                 fv = 6.139136                                                 fv = 6.472837                                                 fv = 5.12684                                                 fv = 6.757808                                                 fv = 1.340781e+154                                                 fv = 5.12684                                                 fv = 5.638732                                                 fv = 7.539059                                                 fv = 1.340781e+154                                                 fv = 5.12684                                                 fv = 5.627676                                                 fv = 5.12684                                                 fv = 5.522092                                                 fv = 5.522092                                                 fv = 5.12684                                                 fv = 6.139136 run # 0: f=5.126839587
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 1: f=1.340780793e+154
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 18.21592                                                 fv = 47.52675                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.672313                                                 fv = 7.672313                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 21.1925                                                 fv = 1.340781e+154                                                 fv = 10.88274                                                 fv = 11.87998                                                 fv = 9.792608                                                 fv = 1.340781e+154                                                 fv = 885.1002                                                 fv = 9.458556                                                 fv = 101689.1                                                 fv = 375648507                                                 fv = 1.340781e+154                                                 fv = 10.30627                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 21.1925                                                 fv = 8.684928                                                 fv = 7.672313                                                 fv = 7.672313                                                 fv = 8.684928                                                 fv = 188147003                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.915949                                                 fv = 8.505663                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.097704                                                 fv = 651.9098                                                 fv = 1.340781e+154                                                 fv = 9.915949                                                 fv = 6.097704                                                 fv = 6.966532                                                 fv = 20.36445                                                 fv = 20.36445                                                 fv = 1.340781e+154                                                 fv = 16.35798                                                 fv = 773.5101                                                 fv = 1.340781e+154                                                 fv = 1495.541                                                 fv = 12.05736                                                 fv = 6.097704                                                 fv = 16.56928                                                 fv = 1.340781e+154                                                 fv = 8.588767                                                 fv = 147.1685                                                 fv = 1.340781e+154                                                 fv = 11.95886                                                 fv = 168924.5                                                 fv = 6.192203                                                 fv = 1.340781e+154                                                 fv = 17.09833                                                 fv = 1.340781e+154                                                 fv = 47.52675                                                 fv = 6.097704                                                 fv = 10.69244                                                 fv = 6.224742                                                 fv = 1.340781e+154                                                 fv = 340.7827                                                 fv = 380.9824                                                 fv = 13.14402                                                 fv = 7.421638                                                 fv = 6.770782                                                 fv = 1.340781e+154                                                 fv = 95.64566                                                 fv = 10.80497                                                 fv = 1229396552                                                 fv = 2362.238                                                 fv = 6.097704                                                 fv = 6.097704                                                 fv = 9.76815                                                 fv = 9.514894                                                 fv = 6.612648                                                 fv = 1.340781e+154                                                 fv = 47.52675                                                 fv = 9.731033                                                 fv = 38.86346                                                 fv = 6.097704                                                 fv = 6.097704                                                 fv = 6.097704                                                 fv = 9.995564                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 367.9261                                                 fv = 12.82967                                                 fv = 7.476777                                                 fv = 6.097704                                                 fv = 6.097704                                                 fv = 8.389928                                                 fv = 7.259887                                                 fv = 6.097704                                                 fv = 7.028777                                                 fv = 6.711268                                                 fv = 12.792                                                 fv = 29.52642 run # 2: f=6.097704308
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.027293                                                 fv = 47.52675                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.510462                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.69185                                                 fv = 207.7579                                                 fv = 10.66028                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 113.5619                                                 fv = 6.510462                                                 fv = 23807019                                                 fv = 6.510462                                                 fv = 1.340781e+154                                                 fv = 10.2189                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 441470.9                                                 fv = 1.340781e+154                                                 fv = 10.2189                                                 fv = 1072654                                                 fv = 10.54234                                                 fv = 11.8579                                                 fv = 194.8042                                                 fv = 8.074154                                                 fv = 28.30585                                                 fv = 9.004389                                                 fv = 127.5468                                                 fv = 51732.04                                                 fv = 18228.5                                                 fv = 1.340781e+154                                                 fv = 32.35111                                                 fv = 6.510462                                                 fv = 6.510462                                                 fv = 254.9314                                                 fv = 7.390994                                                 fv = 6.510462                                                 fv = 6.510462                                                 fv = 6.510462                                                 fv = 19.07817                                                 fv = 254.9314                                                 fv = 388340.9                                                 fv = 23.50146                                                 fv = 12.02038                                                 fv = 11.8579                                                 fv = 7.307451                                                 fv = 8.808058                                                 fv = 7.43319                                                 fv = 9.288853                                                 fv = 77.7568                                                 fv = 8.004778                                                 fv = 90.23597                                                 fv = 97.3029                                                 fv = 14.51278                                                 fv = 6.510462                                                 fv = 6.510462                                                 fv = 6.510462                                                 fv = 19264.39                                                 fv = 6.510462                                                 fv = 6.510462                                                 fv = 7.971148                                                 fv = 16.70169                                                 fv = 6.510462                                                 fv = 10.2091                                                 fv = 10.64681 run # 3: f=6.510462263
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.22486                                                 fv = 47.52675                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.22486                                                 fv = 10.64681                                                 fv = 47.52675                                                 fv = 24.89491                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.110491                                                 fv = 589494.6                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.901097                                                 fv = 1.340781e+154                                                 fv = 1965.688                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2617.253                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.629697                                                 fv = 1.340781e+154                                                 fv = 15.80702                                                 fv = 41.64537                                                 fv = 6.901097                                                 fv = 5.705957                                                 fv = 14.4441                                                 fv = 5.8115e+11                                                 fv = 7.562689                                                 fv = 5.110491                                                 fv = 2073.049                                                 fv = 1.340781e+154                                                 fv = 9.490177                                                 fv = 856.7302                                                 fv = 1.340781e+154                                                 fv = 5.110491                                                 fv = 5.110491                                                 fv = 1.340781e+154                                                 fv = 5.110491                                                 fv = 6.195467                                                 fv = 1.340781e+154                                                 fv = 8.682174                                                 fv = 5.110491                                                 fv = 5.684581                                                 fv = 6.038627                                                 fv = 5.110491                                                 fv = 10.48985                                                 fv = 19460743                                                 fv = 237.7343                                                 fv = 1.340781e+154                                                 fv = 10.82132                                                 fv = 38744175314                                                 fv = 198.6162                                                 fv = 20.36964                                                 fv = 1487.924                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.340276                                                 fv = 5.110491                                                 fv = 7.988812                                                 fv = 7.988812                                                 fv = 8.779379                                                 fv = 7.980024                                                 fv = 10056.68                                                 fv = 7.261816                                                 fv = 16.06155                                                 fv = 5.110491                                                 fv = 5.110491                                                 fv = 5.110491                                                 fv = 5.743919                                                 fv = 5.110491                                                 fv = 1.340781e+154                                                 fv = 5.110491                                                 fv = 8.070203                                                 fv = 1.340781e+154                                                 fv = 5.110491                                                 fv = 9.934461                                                 fv = 7.562689                                                 fv = 5.110491                                                 fv = 5.110491                                                 fv = 1.340781e+154                                                 fv = 11.61817                                                 fv = 20.40843                                                 fv = 1.340781e+154                                                 fv = 5.110491                                                 fv = 5.110491                                                 fv = 5.110491                                                 fv = 7.988812                                                 fv = 5.520074                                                 fv = 5.110491                                                 fv = 6.016774                                                 fv = 5.110491                                                 fv = 6.185119                                                 fv = 5.705957                                                 fv = 7.259569 run # 4: f=5.110491431
#> 
#> bb eval : 642
#> best    : 5.110491431
#> worst   : 1.340780793e+154
#> solution: x = ( 0 0 0 0 0 0 3 0 0 2 1 4 1 1 2 1 1 6 2 1 ) f(x) = 5.110491431
#> 
#>                                                 fv = 5.110491                                                                       Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 70.65463