Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  0  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  3  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  1  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 20.73029                                                 fv = 15.81358                                                 fv = 1.340781e+154                                                 fv = 20.43363                                                 fv = 1.340781e+154                                                 fv = 55.56435                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 136.7153                                                 fv = 15.81358                                                 fv = 1.340781e+154                                                 fv = 3548.72                                                 fv = 55.56435                                                 fv = 1.340781e+154                                                 fv = 46.66694                                                 fv = 46.66694                                                 fv = 1705.005                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.39154                                                 fv = 1.340781e+154                                                 fv = 1346769                                                 fv = 117.8687                                                 fv = 1.340781e+154                                                 fv = 102.3604                                                 fv = 45.72264                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.81358                                                 fv = 15.81358                                                 fv = 128.1016                                                 fv = 529.2198                                                 fv = 6859.36                                                 fv = 44.95156                                                 fv = 1329772                                                 fv = 15.81358                                                 fv = 164.8398                                                 fv = 17.86669                                                 fv = 83.45075                                                 fv = 1.340781e+154                                                 fv = 231.4721                                                 fv = 21.18295                                                 fv = 1.340781e+154                                                 fv = 2105.208                                                 fv = 35.18507                                                 fv = 19.4318                                                 fv = 15.81358                                                 fv = 30.48901                                                 fv = 17.92675                                                 fv = 21.20367                                                 fv = 9.163181                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 529.5741                                                 fv = 7.394791                                                 fv = 68833.1                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.07116                                                 fv = 7.142533                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 529.5741                                                 fv = 28.71159                                                 fv = 44.72265                                                 fv = 8.383693                                                 fv = 12.32269                                                 fv = 8.359488                                                 fv = 25.75688                                                 fv = 8.383693                                                 fv = 8.383693                                                 fv = 8.383693                                                 fv = 1.340781e+154                                                 fv = 19.21231                                                 fv = 6.810184                                                 fv = 95.05722                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1004.169                                                 fv = 47.76864                                                 fv = 19.16448                                                 fv = 6.810184                                                 fv = 9.17611                                                 fv = 35.57127                                                 fv = 6.810184                                                 fv = 6.810184                                                 fv = 10.76594                                                 fv = 6.810184                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.810184                                                 fv = 6.810184                                                 fv = 16.17                                                 fv = 6.818303                                                 fv = 6.734244                                                 fv = 21.18295                                                 fv = 1.340781e+154                                                 fv = 1244.61                                                 fv = 172.1536                                                 fv = 6.710193                                                 fv = 13.46531                                                 fv = 1.340781e+154                                                 fv = 1244.61                                                 fv = 55.6926                                                 fv = 6.466157                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 23.52017                                                 fv = 10.9048                                                 fv = 9.53133                                                 fv = 20.29263                                                 fv = 8.714505                                                 fv = 5.317548                                                 fv = 1.340781e+154                                                 fv = 1875.339                                                 fv = 9.163181                                                 fv = 10.71269                                                 fv = 7.482012                                                 fv = 14.68468                                                 fv = 5.317548                                                 fv = 5.317548                                                 fv = 1.340781e+154                                                 fv = 5.317548                                                 fv = 1.340781e+154                                                 fv = 5.317548                                                 fv = 6.191041                                                 fv = 7.520908                                                 fv = 334.0516                                                 fv = 15.59836                                                 fv = 9.163181                                                 fv = 6.977445                                                 fv = 1.340781e+154                                                 fv = 206.4484                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.319353                                                 fv = 5.317548                                                 fv = 7.523262                                                 fv = 11.32645                                                 fv = 14.01424                                                 fv = 6.084951                                                 fv = 7.482012                                                 fv = 7.520908                                                 fv = 6.424264                                                 fv = 5.317548                                                 fv = 16.68856                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.369016                                                 fv = 7.682938                                                 fv = 6.472456                                                 fv = 8.110659                                                 fv = 5.317548                                                 fv = 5.317548                                                 fv = 5.317548                                                 fv = 12.21773                                                 fv = 8.483827                                                 fv = 8.168765                                                 fv = 9.339097                                                 fv = 8.989396                                                 fv = 13.95172                                                 fv = 2978.237 run # 0: f=5.317548055
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 1: f=1.340780793e+154
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.13462                                                 fv = 36.7919                                                 fv = 10.13462                                                 fv = 17.63177                                                 fv = 360.9569                                                 fv = 8.846142                                                 fv = 16.25961                                                 fv = 25.36278                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 483359.1                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 229.2471                                                 fv = 1.340781e+154                                                 fv = 4873.809                                                 fv = 911.4602                                                 fv = 13.23243                                                 fv = 8.846142                                                 fv = 8.846142                                                 fv = 12.25758                                                 fv = 1.340781e+154                                                 fv = 16.26774                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.06047                                                 fv = 9.733353                                                 fv = 6970.751                                                 fv = 17.31586                                                 fv = 772.1101                                                 fv = 139.9277                                                 fv = 1.340781e+154                                                 fv = 15.87233                                                 fv = 20.44748                                                 fv = 154.5991                                                 fv = 45910.06                                                 fv = 128.257                                                 fv = 128.257                                                 fv = 1.340781e+154                                                 fv = 44.79658                                                 fv = 3389.244                                                 fv = 8.846142                                                 fv = 8.846142                                                 fv = 8.846142                                                 fv = 4330699                                                 fv = 19.20973                                                 fv = 3581130                                                 fv = 10.86982                                                 fv = 34.97127                                                 fv = 43.86724                                                 fv = 28.80303                                                 fv = 1.340781e+154                                                 fv = 9.745862                                                 fv = 16.00506                                                 fv = 31.639                                                 fv = 10.73443                                                 fv = 8.846142                                                 fv = 1.340781e+154                                                 fv = 20.60218                                                 fv = 688.4212                                                 fv = 8.846142                                                 fv = 8.846142                                                 fv = 8.846142                                                 fv = 10.73443                                                 fv = 8.846142                                                 fv = 2002.408                                                 fv = 16.57082 run # 2: f=8.846142347
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 77.61566                                                 fv = 36.7919                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 28.05229                                                 fv = 20.82608                                                 fv = 1034387                                                 fv = 16.97218                                                 fv = 882524.5                                                 fv = 1.340781e+154                                                 fv = 701464363469                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 95512.1                                                 fv = 127.7789                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 76.66386                                                 fv = 2503.359                                                 fv = 9844902404                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 76383.92                                                 fv = 9.615289e+13                                                 fv = 15.53148                                                 fv = 16.75296                                                 fv = 22300.42                                                 fv = 1.340781e+154                                                 fv = 15.53148                                                 fv = 26.03234                                                 fv = 1.340781e+154                                                 fv = 69084415                                                 fv = 17287441658                                                 fv = 1.340781e+154 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.53148                                                 fv = 15.53148                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.53148                                                 fv = 17317.87                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 23.57454                                                 fv = 1.340781e+154                                                 fv = 88.50794                                                 fv = 35.60574                                                 fv = 1.340781e+154                                                 fv = 13.46713                                                 fv = 27864730                                                 fv = 1.340781e+154                                                 fv = 24.66261                                                 fv = 358.9149                                                 fv = 42.76535                                                 fv = 1.340781e+154                                                 fv = 40953750                                                 fv = 1.340781e+154                                                 fv = 13.46713                                                 fv = 1.340781e+154                                                 fv = 17.78335                                                 fv = 13.46713                                                 fv = 18.02338                                                 fv = 13.46713                                                 fv = 1.340781e+154                                                 fv = 13.46713                                                 fv = 13.46713                                                 fv = 22.05592                                                 fv = 23.11908                                                 fv = 20.44744                                                 fv = 14.59115                                                 fv = 14.39757                                                 fv = 12.27978                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 218.5105                                                 fv = 1.340781e+154                                                 fv = 779.1987                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 274.6682                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.27978                                                 fv = 12.27978                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 19.89061                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 63.20096                                                 fv = 34024.85                                                 fv = 12.27978                                                 fv = 13.89961                                                 fv = 15.44819                                                 fv = 1.340781e+154                                                 fv = 10.87764                                                 fv = 1.340781e+154                                                 fv = 155.5837                                                 fv = 15.31916                                                 fv = 1.340781e+154                                                 fv = 12.27978                                                 fv = 18.31996                                                 fv = 25822.21                                                 fv = 8.614915                                                 fv = 24.62111                                                 fv = 1.340781e+154                                                 fv = 75.18248                                                 fv = 9.325514                                                 fv = 9.269641                                                 fv = 14.09366                                                 fv = 12.59601                                                 fv = 18.32486                                                 fv = 10.93796                                                 fv = 10.93796                                                 fv = 8.614915                                                 fv = 5.928139                                                 fv = 69230753                                                 fv = 1.340781e+154                                                 fv = 19.02197                                                 fv = 6.502991                                                 fv = 93805.78                                                 fv = 5.928139                                                 fv = 10.04723                                                 fv = 9.692948                                                 fv = 7.898804                                                 fv = 10.04723                                                 fv = 5.928139                                                 fv = 10.45469                                                 fv = 10.04723                                                 fv = 7.162931                                                 fv = 6.812193                                                 fv = 1.340781e+154                                                 fv = 7.096365                                                 fv = 876.6058                                                 fv = 17.91416                                                 fv = 1.340781e+154                                                 fv = 147.639                                                 fv = 63068.98                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.118794                                                 fv = 22.76894                                                 fv = 21.10573                                                 fv = 9.888819                                                 fv = 24.31075                                                 fv = 27.6367                                                 fv = 1.340781e+154                                                 fv = 5.565935                                                 fv = 1.340781e+154                                                 fv = 28685.39                                                 fv = 10.11188                                                 fv = 21.19293                                                 fv = 9.129542                                                 fv = 5.602371                                                 fv = 289.1715                                                 fv = 5.114392                                                 fv = 1.340781e+154                                                 fv = 20.67548                                                 fv = 1.340781e+154                                                 fv = 5.971206                                                 fv = 6.785186                                                 fv = 45436.76                                                 fv = 9.216443                                                 fv = 7.536946                                                 fv = 5.114392                                                 fv = 23.46079                                                 fv = 27.7603                                                 fv = 5.813487                                                 fv = 8.395593                                                 fv = 5.224435                                                 fv = 5944.152                                                 fv = 5.114392                                                 fv = 6.312439                                                 fv = 1565.502                                                 fv = 9.291981                                                 fv = 6.52682                                                 fv = 7345.558                                                 fv = 1.340781e+154                                                 fv = 12.16074                                                 fv = 4.889013                                                 fv = 4.889013                                                 fv = 1.340781e+154                                                 fv = 12.16074                                                 fv = 5.039655                                                 fv = 4.889013                                                 fv = 4.889013                                                 fv = 69.95912                                                 fv = 7.828868                                                 fv = 4.889013                                                 fv = 4.889013                                                 fv = 5.836108                                                 fv = 6.230427                                                 fv = 4.50888                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.06848                                                 fv = 8.650187                                                 fv = 6.30351                                                 fv = 4.50888                                                 fv = 4.50888                                                 fv = 5.188942                                                 fv = 4.50888                                                 fv = 1.340781e+154                                                 fv = 5.397282                                                 fv = 5.151288                                                 fv = 8.426249                                                 fv = 5.176796                                                 fv = 4.974084                                                 fv = 185.2207                                                 fv = 5.889036                                                 fv = 10.49154                                                 fv = 16.86429                                                 fv = 17.58429                                                 fv = 32.26148                                                 fv = 1.340781e+154                                                 fv = 5.827876                                                 fv = 4.50888                                                 fv = 5.151288                                                 fv = 6.346022                                                 fv = 8.650187                                                 fv = 4.628806                                                 fv = 4.50888                                                 fv = 4.50888                                                 fv = 10.49154                                                 fv = 4.50888                                                 fv = 1.340781e+154                                                 fv = 6.816144                                                 fv = 6.111027                                                 fv = 6.314918                                                 fv = 6.111027 run # 3: f=4.508879831
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 364.3736                                                 fv = 8.527096                                                 fv = 36.7919                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 36.7919                                                 fv = 36.7919                                                 fv = 36.7919                                                 fv = 36.7919                                                 fv = 28.05229                                                 fv = 9.556209                                                 fv = 8.947888                                                 fv = 1.340781e+154                                                 fv = 10.86982                                                 fv = 28.05229                                                 fv = 10.17274                                                 fv = 1650598126                                                 fv = 9.585792                                                 fv = 1.340781e+154                                                 fv = 8.947888                                                 fv = 36.7919                                                 fv = 749505249                                                 fv = 8.527096                                                 fv = 4.941325e+14                                                 fv = 711537393                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 36.7919                                                 fv = 7.122485                                                 fv = 12.08209                                                 fv = 1.340781e+154                                                 fv = 36.7919                                                 fv = 7.122485                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 365.9032                                                 fv = 28.33957                                                 fv = 1.340781e+154                                                 fv = 8.629842                                                 fv = 1.340781e+154                                                 fv = 7.122485                                                 fv = 1.340781e+154                                                 fv = 8.035624                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.122485                                                 fv = 1.340781e+154                                                 fv = 135546.6                                                 fv = 1.340781e+154                                                 fv = 11.22599                                                 fv = 7.122485                                                 fv = 28.33957                                                 fv = 8.041765                                                 fv = 7.049802                                                 fv = 1.340781e+154                                                 fv = 7.049802                                                 fv = 345.6936                                                 fv = 7.049802                                                 fv = 7.049802                                                 fv = 36.7919                                                 fv = 7.049802                                                 fv = 7.049802                                                 fv = 9.571012                                                 fv = 7.049802                                                 fv = 558815.6                                                 fv = 7.049802                                                 fv = 26.78737                                                 fv = 1.340781e+154                                                 fv = 6.473392                                                 fv = 3728485                                                 fv = 1.340781e+154                                                 fv = 148.3469                                                 fv = 50.10099                                                 fv = 1.340781e+154                                                 fv = 9.481071                                                 fv = 9.481071                                                 fv = 15.07231                                                 fv = 6.473392                                                 fv = 7.613393                                                 fv = 5.072627                                                 fv = 1.340781e+154                                                 fv = 104.9538                                                 fv = 4.436137                                                 fv = 5.846278                                                 fv = 1.340781e+154                                                 fv = 3949.607                                                 fv = 34864.58                                                 fv = 8.629842                                                 fv = 34864.58                                                 fv = 124.7511                                                 fv = 1.340781e+154                                                 fv = 5.28909                                                 fv = 4.381296                                                 fv = 1.340781e+154                                                 fv = 18.28642                                                 fv = 11.15074                                                 fv = 1166.661                                                 fv = 7.049802                                                 fv = 1.340781e+154                                                 fv = 4.54149                                                 fv = 4.823572                                                 fv = 7.610664                                                 fv = 5.690731                                                 fv = 6.164599                                                 fv = 1.340781e+154                                                 fv = 16.76932                                                 fv = 48.99369                                                 fv = 34792.57                                                 fv = 5.690731                                                 fv = 20.47411                                                 fv = 1.340781e+154                                                 fv = 24.56755                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.015903                                                 fv = 21.26062                                                 fv = 6.614393                                                 fv = 3.482326                                                 fv = 5.690731                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.613393                                                 fv = 4.381296                                                 fv = 5.690731                                                 fv = 3.482326                                                 fv = 6.614393                                                 fv = 7.021546                                                 fv = 6.218098                                                 fv = 6.593992                                                 fv = 6.218098                                                 fv = 7.049802                                                 fv = 4.381296                                                 fv = 7.129465                                                 fv = 501.6956                                                 fv = 5.690731                                                 fv = 3.482326                                                 fv = 1.340781e+154                                                 fv = 99010.79                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.045943                                                 fv = 3.482326                                                 fv = 4.30322                                                 fv = 5.690731                                                 fv = 5.690731                                                 fv = 3.858161                                                 fv = 3.758374                                                 fv = 3.866872                                                 fv = 3.482326                                                 fv = 7.049802                                                 fv = 4.296578                                                 fv = 1.340781e+154                                                 fv = 3.482326                                                 fv = 5.998848                                                 fv = 3.482326                                                 fv = 3.482326                                                 fv = 7.049802                                                 fv = 7.049802                                                 fv = 7.049802                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.948227                                                 fv = 3.482326                                                 fv = 3.482326                                                 fv = 4.54149                                                 fv = 4.848411                                                 fv = 1.340781e+154                                                 fv = 4.036083                                                 fv = 3.482326                                                 fv = 3.482326                                                 fv = 4.083678                                                 fv = 3.741247                                                 fv = 3.482326                                                 fv = 4.033461                                                 fv = 17.83427                                                 fv = 25.61859                                                 fv = 451.4119                                                 fv = 8.527096                                                 fv = 1.340781e+154                                                 fv = 6.614393                                                 fv = 4.083678                                                 fv = 3.482326                                                 fv = 3.482326                                                 fv = 3.482326                                                 fv = 1.340781e+154                                                 fv = 4.563222                                                 fv = 1.340781e+154                                                 fv = 5.935489                                                 fv = 25.61859                                                 fv = 8.527096 run # 4: f=3.482326243
#> 
#> bb eval : 789
#> best    : 3.482326243
#> worst   : 1.340780793e+154
#> solution: x = ( 0 0 0 0 0 0 1 0 0 2 2 2 1 10 1 1 2 10 1 1 ) f(x) = 3.482326243
#> 
#>                                                 fv = 3.482326                                                                       Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 31.96099