Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  1  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  4  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  2  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 41.06451                                                 fv = 47.29098                                                 fv = 14.63112                                                 fv = 82892426                                                 fv = 1.340781e+154                                                 fv = 29.82809                                                 fv = 33.4317                                                 fv = 253727500                                                 fv = 1.340781e+154                                                 fv = 69.42801                                                 fv = 48.54197                                                 fv = 1.340781e+154                                                 fv = 43.8556                                                 fv = 1.340781e+154                                                 fv = 361.7601                                                 fv = 1.340781e+154                                                 fv = 33.4317                                                 fv = 33.4317                                                 fv = 1.340781e+154                                                 fv = 426228.1                                                 fv = 1.340781e+154                                                 fv = 189.3388                                                 fv = 7566020600                                                 fv = 23862.88                                                 fv = 84.80209                                                 fv = 14.63112                                                 fv = 941.9882                                                 fv = 12711.69                                                 fv = 21.14026                                                 fv = 1.340781e+154                                                 fv = 14.63112                                                 fv = 14.63112                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 38.59956                                                 fv = 30.81262                                                 fv = 722.8917                                                 fv = 364793                                                 fv = 20.10473                                                 fv = 1.340781e+154                                                 fv = 30.81262                                                 fv = 83840.75                                                 fv = 52.78754                                                 fv = 36.866                                                 fv = 5342.981                                                 fv = 1.340781e+154                                                 fv = 11.87166                                                 fv = 1.340781e+154                                                 fv = 13.12137                                                 fv = 11.87166                                                 fv = 20.31869                                                 fv = 13.62149                                                 fv = 1.340781e+154                                                 fv = 11.1145                                                 fv = 180456419                                                 fv = 1419.584                                                 fv = 282.6476                                                 fv = 6.531749                                                 fv = 368.2026                                                 fv = 17.93723                                                 fv = 3248.018                                                 fv = 10.74789                                                 fv = 7.566666                                                 fv = 8.301126                                                 fv = 10.57534                                                 fv = 7.64348                                                 fv = 7.053294                                                 fv = 1.340781e+154                                                 fv = 8.326347                                                 fv = 8.928273                                                 fv = 11.88262                                                 fv = 13.3531                                                 fv = 28.39719                                                 fv = 8.553729                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.669464                                                 fv = 1.340781e+154                                                 fv = 13.96912                                                 fv = 6.531749                                                 fv = 6.112932                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.65539                                                 fv = 6.834391                                                 fv = 8.997378                                                 fv = 6.112932                                                 fv = 14.2242                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.112932                                                 fv = 7.637819                                                 fv = 5.865173                                                 fv = 145.069                                                 fv = 1.340781e+154                                                 fv = 6.412487                                                 fv = 1.340781e+154                                                 fv = 5.865173                                                 fv = 1.340781e+154                                                 fv = 5.865173                                                 fv = 7.420793                                                 fv = 5.865173                                                 fv = 6.704843                                                 fv = 7.972528                                                 fv = 8.648929                                                 fv = 1.340781e+154                                                 fv = 7.859397                                                 fv = 11.24289                                                 fv = 10.65322                                                 fv = 13.75896                                                 fv = 10.98985                                                 fv = 7.753354                                                 fv = 16.95344                                                 fv = 14.47428                                                 fv = 1543579133                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.510637                                                 fv = 7.557987                                                 fv = 5.865173                                                 fv = 7.842776                                                 fv = 7.55813                                                 fv = 8.679324                                                 fv = 1.340781e+154                                                 fv = 8.111549 run # 0: f=5.865173358
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.180806                                                 fv = 36.866                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.82157                                                 fv = 12.11137                                                 fv = 26.32288                                                 fv = 8.348193                                                 fv = 11.36377                                                 fv = 7.180806                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 216050.2                                                 fv = 9.606177                                                 fv = 9.606177                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.180806                                                 fv = 1.340781e+154                                                 fv = 22.33625                                                 fv = 7.180806                                                 fv = 291995237                                                 fv = 7.180806                                                 fv = 1.340781e+154                                                 fv = 11.12722                                                 fv = 6.480398                                                 fv = 792876.2                                                 fv = 12.28242                                                 fv = 14.5617                                                 fv = 14.5617                                                 fv = 35.96231                                                 fv = 136.7599                                                 fv = 47.6174                                                 fv = 5.892681                                                 fv = 187.5277                                                 fv = 1.340781e+154                                                 fv = 31.81034                                                 fv = 487.4265                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1005.071                                                 fv = 10.90183                                                 fv = 13.66744                                                 fv = 7.002766                                                 fv = 14.4161                                                 fv = 517.1615                                                 fv = 1.340781e+154                                                 fv = 8.236081                                                 fv = 14.71825                                                 fv = 6.204771                                                 fv = 8.470588                                                 fv = 14.98457                                                 fv = 1.340781e+154                                                 fv = 6.204771                                                 fv = 8.844322                                                 fv = 1.340781e+154                                                 fv = 1664076                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 43.11659                                                 fv = 48.60104                                                 fv = 9.738385                                                 fv = 5.028069                                                 fv = 5.028069                                                 fv = 1.340781e+154                                                 fv = 8.245658                                                 fv = 6.016404                                                 fv = 1.340781e+154                                                 fv = 24.8372                                                 fv = 1.340781e+154                                                 fv = 7.426907                                                 fv = 9.097429                                                 fv = 7.426907                                                 fv = 5.028069                                                 fv = 9.164639                                                 fv = 7.426907                                                 fv = 6.84831                                                 fv = 8.358127                                                 fv = 7.00411                                                 fv = 5.834193                                                 fv = 327.2945                                                 fv = 5.028069                                                 fv = 5.451355                                                 fv = 34.89592                                                 fv = 1.340781e+154                                                 fv = 5.028069                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.050596                                                 fv = 5.028069                                                 fv = 5.028069                                                 fv = 1.340781e+154                                                 fv = 5.028069                                                 fv = 12.5058                                                 fv = 6.07984                                                 fv = 7.747917                                                 fv = 11.86848                                                 fv = 5.028069                                                 fv = 7.47809                                                 fv = 1.340781e+154                                                 fv = 5.028069                                                 fv = 29.7295                                                 fv = 5.028069                                                 fv = 29.95778                                                 fv = 5.028069                                                 fv = 5.834193                                                 fv = 7.00411                                                 fv = 5.088655                                                 fv = 15.78346                                                 fv = 1.340781e+154                                                 fv = 95.06755                                                 fv = 23.27278                                                 fv = 5.028069                                                 fv = 5.028069                                                 fv = 11.86848                                                 fv = 10.03624                                                 fv = 20.24051                                                 fv = 1.340781e+154                                                 fv = 7.426907                                                 fv = 5.028069                                                 fv = 5.028069                                                 fv = 12.5058                                                 fv = 5.028069                                                 fv = 6.977487                                                 fv = 6.559452 run # 1: f=5.028069317
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.845244                                                 fv = 36.866                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 27.83158                                                 fv = 6.619541                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.619541                                                 fv = 5.233923                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.233923                                                 fv = 368.8469                                                 fv = 46293903                                                 fv = 7.952598                                                 fv = 46528128369                                                 fv = 222229211942                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.233923                                                 fv = 5.233923                                                 fv = 102.0533                                                 fv = 9985.637                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.233923                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 121.7622                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.233923                                                 fv = 5.233923                                                 fv = 13629.41                                                 fv = 5.233923                                                 fv = 6.669                                                 fv = 5.753965                                                 fv = 13.04271                                                 fv = 5.233923                                                 fv = 11870.16                                                 fv = 1.340781e+154                                                 fv = 5.233923                                                 fv = 6.669                                                 fv = 6.669                                                 fv = 1.340781e+154                                                 fv = 255.0858                                                 fv = 2146.375                                                 fv = 6.619541                                                 fv = 1.340781e+154                                                 fv = 15.71323                                                 fv = 12.33328                                                 fv = 12.0341                                                 fv = 5.233923                                                 fv = 5.233923                                                 fv = 5.233923                                                 fv = 15.60735                                                 fv = 16.88612                                                 fv = 1.340781e+154                                                 fv = 18.27105                                                 fv = 5.233923                                                 fv = 1.340781e+154                                                 fv = 5.233923                                                 fv = 5.233923                                                 fv = 7.059972                                                 fv = 11.42919                                                 fv = 6.416081                                                 fv = 6.416081                                                 fv = 1.340781e+154                                                 fv = 11.36886                                                 fv = 12.41025                                                 fv = 83.80615                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.867933                                                 fv = 11.27877                                                 fv = 5.514387                                                 fv = 6.100448                                                 fv = 12.24388                                                 fv = 5.233923                                                 fv = 5.233923                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.233923                                                 fv = 5.233923                                                 fv = 5.233923                                                 fv = 6.060037                                                 fv = 18.06321                                                 fv = 9.334974                                                 fv = 5.781516 run # 2: f=5.233922676
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2822.702                                                 fv = 36.866                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 36.866                                                 fv = 36.866                                                 fv = 45503257                                                 fv = 36.866                                                 fv = 51275857                                                 fv = 456.5791                                                 fv = 35.48428                                                 fv = 658.334                                                 fv = 59.91573                                                 fv = 44.04282                                                 fv = 1.340781e+154                                                 fv = 44.04282                                                 fv = 44.04282                                                 fv = 44.04282                                                 fv = 1.340781e+154                                                 fv = 9.124814                                                 fv = 1.340781e+154                                                 fv = 106.0171                                                 fv = 1.340781e+154                                                 fv = 394.7288                                                 fv = 1.340781e+154                                                 fv = 105.691                                                 fv = 1.340781e+154                                                 fv = 1662.894                                                 fv = 4755.984                                                 fv = 1.340781e+154                                                 fv = 3067.671                                                 fv = 10.7366                                                 fv = 1.340781e+154                                                 fv = 36635.4                                                 fv = 604501.7                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 36.866                                                 fv = 49040.68                                                 fv = 190.2847                                                 fv = 54.15282                                                 fv = 1.340781e+154                                                 fv = 54.15282                                                 fv = 37002.2                                                 fv = 35.75884                                                 fv = 49040.68                                                 fv = 366.3979                                                 fv = 100106825530                                                 fv = 1.340781e+154                                                 fv = 12.28451                                                 fv = 30.47972                                                 fv = 1.340781e+154                                                 fv = 7.51694                                                 fv = 1110.81                                                 fv = 1.340781e+154                                                 fv = 36.866                                                 fv = 8.948823                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.12189                                                 fv = 8.038796                                                 fv = 8.038796                                                 fv = 8.681456                                                 fv = 410701346581                                                 fv = 1.340781e+154                                                 fv = 7.51694                                                 fv = 9.199717                                                 fv = 10.12189                                                 fv = 10.12189                                                 fv = 28.53526                                                 fv = 27154669                                                 fv = 12.64568                                                 fv = 1.340781e+154                                                 fv = 21.8919                                                 fv = 34.98446                                                 fv = 57.16252                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.632984                                                 fv = 17.60387                                                 fv = 11.87912                                                 fv = 62.04234                                                 fv = 1.340781e+154                                                 fv = 7.51694                                                 fv = 1.340781e+154                                                 fv = 7.408349                                                 fv = 6.68181                                                 fv = 1.340781e+154                                                 fv = 36.866                                                 fv = 21926520                                                 fv = 12.09434                                                 fv = 16.23656                                                 fv = 6.68181                                                 fv = 7.641988 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 9.483163                                                 fv = 6.68181                                                 fv = 7.678195                                                 fv = 6.68181                                                 fv = 1.340781e+154                                                 fv = 6.68181                                                 fv = 6.68181                                                 fv = 7.641988 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 19.46755                                                 fv = 479.8733                                                 fv = 1.340781e+154                                                 fv = 29.08849                                                 fv = 137194.6                                                 fv = 7987.057                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 29.08849                                                 fv = 9.699615                                                 fv = 9.699615                                                 fv = 8.427906                                                 fv = 8.427906                                                 fv = 1.340781e+154                                                 fv = 7.32523                                                 fv = 7.678195                                                 fv = 7.678195                                                 fv = 29.95509                                                 fv = 6.68181                                                 fv = 6.68181                                                 fv = 6.68181                                                 fv = 6.68181                                                 fv = 6.68181                                                 fv = 7.678195                                                 fv = 6.98671                                                 fv = 7.678195                                                 fv = 29.08849                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 29.08849                                                 fv = 6.68181                                                 fv = 6.68181                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.68181                                                 fv = 6.68181                                                 fv = 1.340781e+154                                                 fv = 6.68181                                                 fv = 6.68181                                                 fv = 6.68181                                                 fv = 10.65761                                                 fv = 24.10879                                                 fv = 12.98378                                                 fv = 1.340781e+154 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 12.53659                                                 fv = 10.49169                                                 fv = 9.824953                                                 fv = 23.79797                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.24218                                                 fv = 1.340781e+154                                                 fv = 6.409789                                                 fv = 80.58542                                                 fv = 38.79401                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.409789                                                 fv = 6.409789                                                 fv = 7.331028                                                 fv = 6.409789                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.61832                                                 fv = 8.944893                                                 fv = 6.409789                                                 fv = 6.68181                                                 fv = 6.68181                                                 fv = 8.19135                                                 fv = 6.68181                                                 fv = 9.819688                                                 fv = 5.514898                                                 fv = 5.27528                                                 fv = 38.21341                                                 fv = 15.14239                                                 fv = 5.27528                                                 fv = 5.27528                                                 fv = 5.27528                                                 fv = 5.27528                                                 fv = 5.27528                                                 fv = 5.27528                                                 fv = 6.770592                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.27528                                                 fv = 5.27528                                                 fv = 1.340781e+154                                                 fv = 5.514898                                                 fv = 5.469233                                                 fv = 5.514898                                                 fv = 5.348743                                                 fv = 5.27528                                                 fv = 7.343463                                                 fv = 6.587398                                                 fv = 6.049444                                                 fv = 8.454513                                                 fv = 5.27528                                                 fv = 1.340781e+154                                                 fv = 7.14499                                                 fv = 5.27528                                                 fv = 17.01703                                                 fv = 10.64092                                                 fv = 20.64655                                                 fv = 36.866                                                 fv = 5.431191                                                 fv = 5.431191                                                 fv = 6.418978                                                 fv = 6.601008                                                 fv = 5.27528                                                 fv = 5.27528                                                 fv = 1.340781e+154                                                 fv = 5.459023                                                 fv = 6.564225                                                 fv = 5.27528                                                 fv = 5.148611                                                 fv = 743.7497                                                 fv = 1.340781e+154                                                 fv = 36.866                                                 fv = 5.148611                                                 fv = 24.81163                                                 fv = 5.148611                                                 fv = 14.63764                                                 fv = 10.50216                                                 fv = 5.148611                                                 fv = 8.794781                                                 fv = 5.148611                                                 fv = 11.61918                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.508868                                                 fv = 6.701923                                                 fv = 12.27419                                                 fv = 5.148611                                                 fv = 5.148611                                                 fv = 5.148611                                                 fv = 9.867644                                                 fv = 8.664706                                                 fv = 22397.92 run # 3: f=5.148610584
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 17.92283                                                 fv = 36.866                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 17.92283                                                 fv = 8.19135                                                 fv = 29.08849                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.19135                                                 fv = 1.340781e+154                                                 fv = 29.08849                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.50332                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2051462303                                                 fv = 634168698051                                                 fv = 11.85473                                                 fv = 1.340781e+154                                                 fv = 17.26547                                                 fv = 8.19135                                                 fv = 4686719                                                 fv = 10.21818                                                 fv = 737.1269                                                 fv = 12.82157                                                 fv = 39.92211                                                 fv = 6.68181                                                 fv = 29.08849                                                 fv = 32.17058                                                 fv = 29.08849                                                 fv = 8.474687                                                 fv = 6.68181                                                 fv = 9.534078                                                 fv = 14.64037                                                 fv = 6.68181                                                 fv = 7.408349                                                 fv = 18344.48                                                 fv = 829053.4                                                 fv = 8.695594                                                 fv = 139.1717                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.68181                                                 fv = 6.701923                                                 fv = 9.867644                                                 fv = 737.1269                                                 fv = 41.86182                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 13635.1                                                 fv = 16.54036                                                 fv = 9.483297                                                 fv = 8.41059                                                 fv = 10.28142                                                 fv = 8.19135                                                 fv = 14.50332                                                 fv = 22.74731                                                 fv = 6.68181                                                 fv = 1.340781e+154                                                 fv = 6.98671                                                 fv = 9.824953                                                 fv = 11.85473                                                 fv = 10.80654                                                 fv = 10.80654                                                 fv = 1.340781e+154                                                 fv = 9.483297                                                 fv = 9.483297                                                 fv = 8.19135                                                 fv = 8.19135                                                 fv = 7.67797                                                 fv = 17.04559                                                 fv = 70.83241                                                 fv = 28.99562                                                 fv = 7.641988                                                 fv = 7.32523                                                 fv = 6.68181                                                 fv = 6.68181                                                 fv = 6.68181                                                 fv = 7.817027                                                 fv = 8.238894                                                 fv = 6.409789                                                 fv = 80.58542                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.409789                                                 fv = 5.514898                                                 fv = 5.27528                                                 fv = 5.27528                                                 fv = 1.340781e+154                                                 fv = 29.08849                                                 fv = 10.85957                                                 fv = 5.27528                                                 fv = 7.51694                                                 fv = 6.564225                                                 fv = 5.27528                                                 fv = 5.514898                                                 fv = 5.514898                                                 fv = 5.27528                                                 fv = 5.514898                                                 fv = 6.418978                                                 fv = 5.362969                                                 fv = 9.287632                                                 fv = 5.27528                                                 fv = 4.395439                                                 fv = 58.83709                                                 fv = 1.340781e+154                                                 fv = 36.866                                                 fv = 12.30804                                                 fv = 9.310292                                                 fv = 16.19804                                                 fv = 7.621999                                                 fv = 6.578327                                                 fv = 4.395439                                                 fv = 9.275104                                                 fv = 4.395439                                                 fv = 9.315886                                                 fv = 8.160046                                                 fv = 8.589158                                                 fv = 5.27528                                                 fv = 7.331028                                                 fv = 13.32326                                                 fv = 11.64872                                                 fv = 9.310292                                                 fv = 11.64872                                                 fv = 843.6856                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 36.866                                                 fv = 7.32523                                                 fv = 4.395439                                                 fv = 4.395439                                                 fv = 9.684403                                                 fv = 1.340781e+154                                                 fv = 4.395439                                                 fv = 4.395439                                                 fv = 9.292813                                                 fv = 4.395439                                                 fv = 7.621999                                                 fv = 8.160046                                                 fv = 9.310292                                                 fv = 9.310292                                                 fv = 9.310292                                                 fv = 9.315886                                                 fv = 16.38971                                                 fv = 10.19806                                                 fv = 1.340781e+154                                                 fv = 8.803544                                                 fv = 1.340781e+154                                                 fv = 4.395439                                                 fv = 4.395439                                                 fv = 6.578327                                                 fv = 20.82715                                                 fv = 4.395439                                                 fv = 4.395439                                                 fv = 4.395439                                                 fv = 9.275104                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.64872                                                 fv = 9.867644 run # 4: f=4.395438842
#> 
#> bb eval : 844
#> best    : 4.395438842
#> worst   : 5.865173358
#> solution: x = ( 0 0 0 0 0 0 0 1 0 2 1 2 6 1 1 1 2 3 5 1 ) f(x) = 4.395438842
#> 
#>                                                 fv = 4.395439                                                                       Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 9.007326