Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  0  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  2  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  1  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 14.78504                                                 fv = 58.25096                                                 fv = 21.9609                                                 fv = 35.35409                                                 fv = 739.4883                                                 fv = 11.15799                                                 fv = 33090.51                                                 fv = 1.340781e+154                                                 fv = 16.85008                                                 fv = 17.66802                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.467135                                                 fv = 8.467135                                                 fv = 8.467135                                                 fv = 18.3267                                                 fv = 5882.283                                                 fv = 197905.2                                                 fv = 19.04262                                                 fv = 1.340781e+154                                                 fv = 8.821561                                                 fv = 8.467135                                                 fv = 7.823205                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.823205                                                 fv = 18.67836                                                 fv = 1.340781e+154                                                 fv = 19.57024                                                 fv = 7.823205                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4091.575                                                 fv = 7924544                                                 fv = 9.235529                                                 fv = 1341.498                                                 fv = 9.206245                                                 fv = 11.32696                                                 fv = 1.340781e+154                                                 fv = 17.012                                                 fv = 12.13007                                                 fv = 9.235529                                                 fv = 9.083101                                                 fv = 9.083101                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.901865                                                 fv = 7.823205                                                 fv = 11.31294                                                 fv = 8.202821                                                 fv = 9.531661                                                 fv = 9.531661                                                 fv = 12.22492                                                 fv = 8.358493                                                 fv = 8.124667                                                 fv = 1.340781e+154                                                 fv = 7.863491                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.17885                                                 fv = 41.16479                                                 fv = 1.340781e+154                                                 fv = 9.879883                                                 fv = 11.9343                                                 fv = 7.823205                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.230759                                                 fv = 7.823205                                                 fv = 7.823205                                                 fv = 7.823205                                                 fv = 7.823205                                                 fv = 1.340781e+154                                                 fv = 7.823205                                                 fv = 7.823205                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.440706                                                 fv = 11.12344                                                 fv = 7.823205                                                 fv = 12.275                                                 fv = 14.44021                                                 fv = 7.946056                                                 fv = 7.495044                                                 fv = 8.049281                                                 fv = 25.80108                                                 fv = 7.064194                                                 fv = 8.951736                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.032669                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.873095                                                 fv = 7.290478                                                 fv = 10.56826                                                 fv = 7.064194                                                 fv = 7.064194                                                 fv = 9.338167                                                 fv = 18.03474                                                 fv = 7.078917                                                 fv = 7.064194                                                 fv = 14.21371                                                 fv = 7.064194                                                 fv = 11.19835                                                 fv = 7.681034                                                 fv = 1.340781e+154                                                 fv = 8.448142                                                 fv = 1.340781e+154                                                 fv = 19.43046                                                 fv = 1776117                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.92058                                                 fv = 7.064194                                                 fv = 6.412205                                                 fv = 10.67926                                                 fv = 1.340781e+154                                                 fv = 8.951736                                                 fv = 6.302119                                                 fv = 6.302119                                                 fv = 1.340781e+154                                                 fv = 6.302119                                                 fv = 6.302119                                                 fv = 6.302119                                                 fv = 6.302119                                                 fv = 1.340781e+154                                                 fv = 7.863491                                                 fv = 6.302119                                                 fv = 7.189038                                                 fv = 6.416101                                                 fv = 6.761616                                                 fv = 6.473267                                                 fv = 7.064194                                                 fv = 7.078917                                                 fv = 8.188731                                                 fv = 9.012347 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 12.58019                                                 fv = 12.57329                                                 fv = 7.773126                                                 fv = 1.340781e+154                                                 fv = 2797.832                                                 fv = 14.40107                                                 fv = 1.340781e+154                                                 fv = 10.65491                                                 fv = 8.425836                                                 fv = 7.687862                                                 fv = 6.628435                                                 fv = 6.302119                                                 fv = 6.416101                                                 fv = 1.340781e+154                                                 fv = 6.82715                                                 fv = 5.844499                                                 fv = 19.7419                                                 fv = 1.340781e+154                                                 fv = 19.28314                                                 fv = 7.314409                                                 fv = 5.844499                                                 fv = 6.85718                                                 fv = 5.844499                                                 fv = 5.844499                                                 fv = 5.844499                                                 fv = 5.844499                                                 fv = 5.844499                                                 fv = 6.518007                                                 fv = 5.537251                                                 fv = 1.340781e+154                                                 fv = 8.045265                                                 fv = 5.537251                                                 fv = 5.537251                                                 fv = 5.537251                                                 fv = 5.537251                                                 fv = 5.537251                                                 fv = 6.263315                                                 fv = 5.402472                                                 fv = 17.2514                                                 fv = 1.340781e+154                                                 fv = 7.677542                                                 fv = 5.402472                                                 fv = 5.402472                                                 fv = 5.402472                                                 fv = 5.364245                                                 fv = 7.688206                                                 fv = 1.340781e+154                                                 fv = 16.03107                                                 fv = 20.57515                                                 fv = 5.741567                                                 fv = 5.562138                                                 fv = 9.585594                                                 fv = 6.363896                                                 fv = 8.01823                                                 fv = 7.149261                                                 fv = 11.70548                                                 fv = 5.741567                                                 fv = 1.340781e+154                                                 fv = 5.364245                                                 fv = 6.657711                                                 fv = 6.32598                                                 fv = 5.364245                                                 fv = 5.364245                                                 fv = 5.613555                                                 fv = 6.626078                                                 fv = 5.514682                                                 fv = 5.579                                                 fv = 48.2389                                                 fv = 1.340781e+154                                                 fv = 6.385437                                                 fv = 6.871103                                                 fv = 5.364245                                                 fv = 5.364245                                                 fv = 5.364245                                                 fv = 5.364245                                                 fv = 5.364245                                                 fv = 6.184038                                                 fv = 7.020508                                                 fv = 6.452002                                                 fv = 7.934504                                                 fv = 6.054475                                                 fv = 11.03302                                                 fv = 8.030933                                                 fv = 6.552861                                                 fv = 13.16                                                 fv = 39.47741                                                 fv = 5.364245 run # 0: f=5.364244636
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 83.18927                                                 fv = 21.67149                                                 fv = 40.4228                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.634662                                                 fv = 10.00696                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 13.49215                                                 fv = 8.73635                                                 fv = 8.859335                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.644874                                                 fv = 120.4054                                                 fv = 9.014406                                                 fv = 58.14616                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 48.4745                                                 fv = 8556.327                                                 fv = 171607.1                                                 fv = 81.26402                                                 fv = 11.76683                                                 fv = 126.0599                                                 fv = 24.41049                                                 fv = 1.340781e+154                                                 fv = 95959.1                                                 fv = 9.206501                                                 fv = 10.21577                                                 fv = 1.340781e+154                                                 fv = 23.46589                                                 fv = 48.4745                                                 fv = 8.644874                                                 fv = 8.573461                                                 fv = 122.0623                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1124.305                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 281535.1                                                 fv = 281535.1                                                 fv = 6131432                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.36402                                                 fv = 8.634645                                                 fv = 13.14273                                                 fv = 8.573461                                                 fv = 8.573461                                                 fv = 13940.23                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.634662                                                 fv = 278.8389                                                 fv = 43.21214                                                 fv = 149009.2                                                 fv = 82.17812                                                 fv = 1.340781e+154                                                 fv = 11.21224                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.573461                                                 fv = 11.21224                                                 fv = 12.11448                                                 fv = 8.964441                                                 fv = 9.348837                                                 fv = 6.105638                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 36.05576                                                 fv = 12.68653                                                 fv = 4.211778                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 889.0076                                                 fv = 1.340781e+154                                                 fv = 23.71717                                                 fv = 52.2703                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 29.44484                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 39.87045                                                 fv = 2159.25                                                 fv = 1.340781e+154                                                 fv = 12.68653                                                 fv = 1.340781e+154                                                 fv = 278.8036                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 889.0076                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.24571                                                 fv = 33614.89                                                 fv = 33614.89                                                 fv = 331.6512                                                 fv = 1.340781e+154                                                 fv = 33614.89                                                 fv = 602862.6                                                 fv = 1.340781e+154                                                 fv = 17016.61                                                 fv = 19.72457                                                 fv = 1.340781e+154                                                 fv = 49.2802                                                 fv = 1.340781e+154                                                 fv = 40.16314                                                 fv = 8.330129                                                 fv = 15.58708                                                 fv = 1.340781e+154                                                 fv = 19.48807                                                 fv = 10.62969                                                 fv = 17.32553                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.96259                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.211778                                                 fv = 184.0583                                                 fv = 23.82324                                                 fv = 4.211778                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.17819                                                 fv = 6.984767                                                 fv = 19.37257                                                 fv = 8235.435                                                 fv = 4.211778                                                 fv = 1.340781e+154                                                 fv = 380.2402                                                 fv = 19.75914                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.877767                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 55134.77                                                 fv = 2.000774                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.84883                                                 fv = 10060.34                                                 fv = 38.9216                                                 fv = 1.340781e+154                                                 fv = 2.000774                                                 fv = 1.340781e+154                                                 fv = 23.00896                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2.000774                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 40.25109                                                 fv = 11.49922                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 125.5855                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 40.39463                                                 fv = 2.000774                                                 fv = 2.000774                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 52.59806                                                 fv = 1.340781e+154                                                 fv = 2.000774                                                 fv = 43.28635                                                 fv = 2.000774                                                 fv = 1.340781e+154                                                 fv = 36.75005                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 1: f=2.000774138
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.765589                                                 fv = 40.4228                                                 fv = 1.340781e+154                                                 fv = 7.620999                                                 fv = 9.013307                                                 fv = 10.61139                                                 fv = 128.4698                                                 fv = 122.8889                                                 fv = 6.765589                                                 fv = 1981989                                                 fv = 128.4698                                                 fv = 32.60129                                                 fv = 1.340781e+154                                                 fv = 174.9069                                                 fv = 128.4698                                                 fv = 1.340781e+154                                                 fv = 13.75249                                                 fv = 1.340781e+154                                                 fv = 85.67455                                                 fv = 407.7492                                                 fv = 1.340781e+154                                                 fv = 19.45117                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.25548                                                 fv = 13.25028                                                 fv = 45.27473                                                 fv = 37.407                                                 fv = 6.765589                                                 fv = 6.765589                                                 fv = 15.26599                                                 fv = 38.56164                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7869.77                                                 fv = 6.765589                                                 fv = 106.9999                                                 fv = 6.765589                                                 fv = 14.55622                                                 fv = 210.5436                                                 fv = 1.340781e+154                                                 fv = 14.55622                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 21.2819                                                 fv = 28.07146                                                 fv = 8.741256                                                 fv = 9.436343                                                 fv = 6.765589                                                 fv = 1.340781e+154                                                 fv = 11.00513                                                 fv = 16.19507                                                 fv = 34.79252                                                 fv = 6.765589                                                 fv = 22.31435                                                 fv = 42.92894                                                 fv = 6.765589                                                 fv = 6.765589                                                 fv = 12.9791                                                 fv = 9.299022                                                 fv = 6.765589                                                 fv = 6.765589                                                 fv = 1.340781e+154                                                 fv = 9.966319 run # 2: f=6.765589176
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 571.1391                                                 fv = 72.58221                                                 fv = 40.4228                                                 fv = 10.00696                                                 fv = 10.00839                                                 fv = 16.8908                                                 fv = 7.372392                                                 fv = 15.5623                                                 fv = 7.372392                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.00487                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.0015                                                 fv = 39.61366                                                 fv = 15.58282                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.34164                                                 fv = 1.340781e+154                                                 fv = 8.670582                                                 fv = 7.367718                                                 fv = 18.19217                                                 fv = 10.00696                                                 fv = 94.68914                                                 fv = 212.7671                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.5623                                                 fv = 13.33419                                                 fv = 266.1327                                                 fv = 11.9589                                                 fv = 29.50314                                                 fv = 13.33419                                                 fv = 13.33419                                                 fv = 3569263                                                 fv = 925.6982                                                 fv = 116414.9                                                 fv = 9.200801                                                 fv = 1.340781e+154                                                 fv = 7.367718                                                 fv = 1.340781e+154                                                 fv = 7.367718                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.00696                                                 fv = 28.11279                                                 fv = 15.40385                                                 fv = 1.340781e+154                                                 fv = 11.54366                                                 fv = 1.340781e+154                                                 fv = 13.33419                                                 fv = 14.4554                                                 fv = 11.9589                                                 fv = 11.9589                                                 fv = 11.9589                                                 fv = 1.340781e+154                                                 fv = 50.98157                                                 fv = 7.367718                                                 fv = 6.628435                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.473267                                                 fv = 10.36304                                                 fv = 1.340781e+154                                                 fv = 12.79852                                                 fv = 11.62807                                                 fv = 22.91228                                                 fv = 1150801                                                 fv = 6.473267                                                 fv = 6.473267                                                 fv = 6.473267                                                 fv = 9.634662                                                 fv = 1.340781e+154                                                 fv = 5.561401                                                 fv = 127.6814                                                 fv = 1.340781e+154                                                 fv = 16.83613                                                 fv = 5.561401                                                 fv = 10.40404                                                 fv = 10.7458                                                 fv = 6.478473                                                 fv = 10.7458                                                 fv = 6.628435                                                 fv = 6.478473                                                 fv = 6.478473                                                 fv = 8.886793                                                 fv = 8.886793                                                 fv = 6.478473                                                 fv = 13.68239                                                 fv = 32.33337                                                 fv = 9.285881                                                 fv = 583.8884                                                 fv = 139.3953                                                 fv = 1.340781e+154                                                 fv = 8519.045                                                 fv = 9.768112                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.189958                                                 fv = 63.06555                                                 fv = 5.561401                                                 fv = 5.561401                                                 fv = 8.633964                                                 fv = 5.561401                                                 fv = 6.929095                                                 fv = 5.561401                                                 fv = 5.561401                                                 fv = 1.340781e+154                                                 fv = 9.77159                                                 fv = 5.561401                                                 fv = 16.93016                                                 fv = 5.561401                                                 fv = 6.735092                                                 fv = 14.8357                                                 fv = 7.045767                                                 fv = 9.494854                                                 fv = 1.340781e+154                                                 fv = 10.94064                                                 fv = 5.561401                                                 fv = 15.30532                                                 fv = 1.340781e+154                                                 fv = 33.30877                                                 fv = 8.886793                                                 fv = 10.46254                                                 fv = 9.359092                                                 fv = 15.54746                                                 fv = 4.312212                                                 fv = 53.68958                                                 fv = 1.340781e+154                                                 fv = 9.429956                                                 fv = 4.312212                                                 fv = 10.1923                                                 fv = 11.05615                                                 fv = 61.24988                                                 fv = 6.61843                                                 fv = 12.31001                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.11639                                                 fv = 62.96312                                                 fv = 4.312212                                                 fv = 4.312212                                                 fv = 9.77159                                                 fv = 9.77159                                                 fv = 8.633964                                                 fv = 4.312212                                                 fv = 4.312212                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.634662                                                 fv = 13.20651                                                 fv = 29.9176                                                 fv = 13.20651                                                 fv = 10.91892                                                 fv = 11.9991                                                 fv = 9.627608                                                 fv = 29.69305                                                 fv = 4.312212                                                 fv = 4.312212                                                 fv = 4.312212                                                 fv = 21.67503                                                 fv = 4.312212                                                 fv = 31.09412                                                 fv = 16.16824                                                 fv = 83.99754                                                 fv = 11.55959 run # 3: f=4.312212064
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 4: f=1.340780793e+154
#> 
#> bb eval : 769
#> best    : 2.000774138
#> worst   : 1.340780793e+154
#> solution: x = ( 0 0 1 9 1 0 0 1 0 4 1 10 1 1 1 6 10 4 1 1 ) f(x) = 2.000774138
#> 
#>                                                 fv = 2.000774                                                                       Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 44.60217