Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  1  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  4  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  2  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 15.46632                                                 fv = 44.29554                                                 fv = 16.88581                                                 fv = 586.1971                                                 fv = 713.2336                                                 fv = 11.66766                                                 fv = 219.5782                                                 fv = 1.340781e+154                                                 fv = 19.2463                                                 fv = 23.41807                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.86083                                                 fv = 8.599713                                                 fv = 39.005                                                 fv = 1.340781e+154                                                 fv = 38.26389                                                 fv = 25.72227                                                 fv = 1.340781e+154                                                 fv = 17.26203                                                 fv = 12.57338                                                 fv = 17.91685                                                 fv = 47.29964                                                 fv = 1.340781e+154                                                 fv = 47.29964                                                 fv = 1.86897e+11                                                 fv = 288037948                                                 fv = 287.2508                                                 fv = 8.599713                                                 fv = 1.340781e+154                                                 fv = 12.14745                                                 fv = 1.340781e+154                                                 fv = 1341.675                                                 fv = 80.61601                                                 fv = 10.17933                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 14.46873                                                 fv = 1.340781e+154                                                 fv = 6146.771                                                 fv = 172081801530                                                 fv = 9.025633                                                 fv = 35.14129                                                 fv = 54.94999                                                 fv = 22.5737                                                 fv = 38.88866                                                 fv = 27582.48                                                 fv = 15.34248                                                 fv = 1.340781e+154                                                 fv = 8.599713                                                 fv = 1.340781e+154                                                 fv = 10.28984                                                 fv = 56.53448                                                 fv = 8.599713                                                 fv = 1.340781e+154                                                 fv = 5956.224                                                 fv = 9.101422                                                 fv = 1.340781e+154                                                 fv = 6.573093                                                 fv = 9.464254                                                 fv = 1.340781e+154                                                 fv = 17.73906                                                 fv = 1.340781e+154                                                 fv = 17.74749                                                 fv = 1.340781e+154                                                 fv = 14.914                                                 fv = 6.509453                                                 fv = 3.593138                                                 fv = 8.256079                                                 fv = 1.340781e+154                                                 fv = 13.26709                                                 fv = 9301.284                                                 fv = 345.3852                                                 fv = 3.593138                                                 fv = 4.288694                                                 fv = 5.334497                                                 fv = 29805.31                                                 fv = 942.1527                                                 fv = 193781.9                                                 fv = 3.239283                                                 fv = 1.340781e+154                                                 fv = 35.39762                                                 fv = 1.340781e+154                                                 fv = 76.49269                                                 fv = 45.60884                                                 fv = 3.943131                                                 fv = 3.351782                                                 fv = 3.431388                                                 fv = 6.292567                                                 fv = 13.51759                                                 fv = 3.397904                                                 fv = 5.816046                                                 fv = 9.131604                                                 fv = 3.239283                                                 fv = 3.239283                                                 fv = 24.49978                                                 fv = 3.397904                                                 fv = 3.397904                                                 fv = 1.340781e+154                                                 fv = 4.429823                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.349319                                                 fv = 20.61951                                                 fv = 1.666043                                                 fv = 128699264                                                 fv = 1.340781e+154                                                 fv = 7.816547                                                 fv = 2.346868                                                 fv = 2.496761                                                 fv = 2.341175                                                 fv = 2.911391                                                 fv = 2.911391                                                 fv = 1.510565                                                 fv = 1.340781e+154                                                 fv = 2718.162                                                 fv = 14.9166                                                 fv = 1.510565                                                 fv = 1.564313                                                 fv = 8.974989                                                 fv = 6.999559                                                 fv = 6.834215                                                 fv = 4.036316                                                 fv = 1.510565                                                 fv = 1.510565                                                 fv = 19.3978                                                 fv = 1.750784                                                 fv = 3.659827                                                 fv = 9.744068                                                 fv = 12.3778                                                 fv = 2.652069                                                 fv = 27.41228                                                 fv = 7.678761                                                 fv = 73.40151                                                 fv = 38.79906                                                 fv = 91.79876                                                 fv = 1.510565                                                 fv = 171228.8                                                 fv = 1.340781e+154                                                 fv = 770.2729                                                 fv = 31.68589                                                 fv = 1.510565                                                 fv = 2.522779                                                 fv = 1.975589                                                 fv = 2.066636                                                 fv = 2.171602                                                 fv = 28.71631                                                 fv = 1.927509                                                 fv = 1.510565                                                 fv = 1.510565                                                 fv = 1.663837                                                 fv = 1.86219                                                 fv = 1.813953                                                 fv = 9.129526                                                 fv = 1.510565                                                 fv = 1332.374                                                 fv = 7.501301                                                 fv = 229.869                                                 fv = 1.340781e+154                                                 fv = 4801.773                                                 fv = 21.7111                                                 fv = 1.637581                                                 fv = 2.223868                                                 fv = 7.619102                                                 fv = 1.510565                                                 fv = 6.726238                                                 fv = 5.693789                                                 fv = 2.791278                                                 fv = 1.71016                                                 fv = 13.82565                                                 fv = 2.751241                                                 fv = 2.259608                                                 fv = 1.643434                                                 fv = 15.51947                                                 fv = 1.585548                                                 fv = 1.340781e+154                                                 fv = 3.118383                                                 fv = 8.623214                                                 fv = 6.726238                                                 fv = 2.066636                                                 fv = 5.978235                                                 fv = 2.060757 run # 0: f=1.510565197
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14736062                                                 fv = 29.8821                                                 fv = 29.8821                                                 fv = 29.8821                                                 fv = 29.8821                                                 fv = 39.58142                                                 fv = 68362645                                                 fv = 14.00542                                                 fv = 1359.535                                                 fv = 2616.42                                                 fv = 133.9977                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16785143                                                 fv = 3.722589                                                 fv = 20527837                                                 fv = 3.700679                                                 fv = 3.37833                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3.890709                                                 fv = 3.890709                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.754259                                                 fv = 3.890709                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.34987                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.01927                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3.209817                                                 fv = 4.741026                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3.209817                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 39.38592                                                 fv = 4.741026                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3.209817                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3.209817                                                 fv = 3.209817                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 26.81697                                                 fv = 90.47847                                                 fv = 1.340781e+154                                                 fv = 50.05343                                                 fv = 3.209817                                                 fv = 905.3856                                                 fv = 1.340781e+154                                                 fv = 3.209817                                                 fv = 5.290536                                                 fv = 3.890709                                                 fv = 2.85288                                                 fv = 2.85288                                                 fv = 1.340781e+154                                                 fv = 328.02                                                 fv = 15.08914                                                 fv = 1.340781e+154                                                 fv = 3.423129                                                 fv = 3.423129                                                 fv = 2.85288                                                 fv = 1.340781e+154                                                 fv = 25.18406                                                 fv = 4.231086                                                 fv = 113584                                                 fv = 2.85288                                                 fv = 1.340781e+154                                                 fv = 2.85288                                                 fv = 4.231086                                                 fv = 80.88618                                                 fv = 4.231086                                                 fv = 2.85288                                                 fv = 4.231086                                                 fv = 1.340781e+154                                                 fv = 2.892475                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3.37833                                                 fv = 642208                                                 fv = 17.35399                                                 fv = 2.85288                                                 fv = 1789.657                                                 fv = 2.85288                                                 fv = 2.85288                                                 fv = 677078.1                                                 fv = 1.340781e+154                                                 fv = 4.908402                                                 fv = 2.85288                                                 fv = 2.85288                                                 fv = 4.754259                                                 fv = 5.406363                                                 fv = 3.890709                                                 fv = 325039.6                                                 fv = 156.1197                                                 fv = 3102.593                                                 fv = 1.340781e+154                                                 fv = 92514.46                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 108.5353                                                 fv = 3.37833                                                 fv = 2.85288                                                 fv = 2.976758                                                 fv = 3.443991                                                 fv = 3.423129                                                 fv = 3.982497                                                 fv = 1.340781e+154                                                 fv = 540.0067                                                 fv = 583.3914                                                 fv = 2.85288                                                 fv = 1.340781e+154                                                 fv = 2.85288                                                 fv = 2.85288                                                 fv = 2.85288                                                 fv = 2.85288                                                 fv = 7.287846                                                 fv = 31.43064                                                 fv = 23.46529                                                 fv = 255.0046                                                 fv = 29.8821                                                 fv = 1.340781e+154                                                 fv = 108.5353                                                 fv = 4.754259                                                 fv = 3.264843                                                 fv = 3.890709                                                 fv = 3.126348                                                 fv = 3.126348                                                 fv = 3.847157                                                 fv = 2.85288                                                 fv = 2.85288                                                 fv = 2.85288                                                 fv = 2.85288                                                 fv = 2.85288                                                 fv = 8.112775                                                 fv = 2.85288                                                 fv = 2.976758                                                 fv = 2.447436                                                 fv = 1.903263                                                 fv = 5.387877                                                 fv = 305.1447                                                 fv = 2.895064                                                 fv = 2.994971                                                 fv = 2.533277                                                 fv = 1.903263                                                 fv = 1.903263                                                 fv = 3.612532                                                 fv = 1.903263                                                 fv = 1.903263                                                 fv = 2.377911                                                 fv = 1.903263                                                 fv = 2.177153                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.45889                                                 fv = 1.903263                                                 fv = 5.198569                                                 fv = 5.744066                                                 fv = 5.0675                                                 fv = 6.175675                                                 fv = 5.198569                                                 fv = 3400282                                                 fv = 1.340781e+154                                                 fv = 14169.61                                                 fv = 2.270255                                                 fv = 1.976077                                                 fv = 1.903263                                                 fv = 1.903263                                                 fv = 1.903263                                                 fv = 3.344191                                                 fv = 2.274223                                                 fv = 2.111049                                                 fv = 2.015376                                                 fv = 4.435913 run # 1: f=1.903262609
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 917.664                                                 fv = 29.8821                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 23.55271                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.29573                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 482.8708                                                 fv = 10.53821                                                 fv = 10.9555                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.08377                                                 fv = 12.64699                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 44751791980                                                 fv = 1.340781e+154                                                 fv = 402177.1                                                 fv = 1.340781e+154                                                 fv = 12.30959                                                 fv = 1.340781e+154                                                 fv = 21.81256                                                 fv = 12.9569                                                 fv = 5.309363                                                 fv = 1.340781e+154                                                 fv = 67.54699                                                 fv = 6730247693                                                 fv = 1.340781e+154                                                 fv = 5.617776                                                 fv = 3.468899                                                 fv = 6.244131                                                 fv = 8.154732                                                 fv = 10.62262                                                 fv = 1.340781e+154                                                 fv = 10.44276                                                 fv = 10.44276                                                 fv = 41.9182                                                 fv = 365.6157                                                 fv = 36.61665                                                 fv = 1.340781e+154                                                 fv = 285021003                                                 fv = 7.092519                                                 fv = 4.495079                                                 fv = 1.340781e+154                                                 fv = 3.569182                                                 fv = 4.890219                                                 fv = 3.569182                                                 fv = 5.905028                                                 fv = 1.340781e+154                                                 fv = 28.96946                                                 fv = 9.598746                                                 fv = 278079267                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6717.292                                                 fv = 3.725705                                                 fv = 2.380387e+13                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.244131                                                 fv = 4.144331                                                 fv = 4.371268                                                 fv = 4.356623                                                 fv = 3.468899                                                 fv = 3.468899                                                 fv = 12.87131                                                 fv = 9.598746                                                 fv = 2.949878                                                 fv = 1.340781e+154                                                 fv = 26.44592                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.077801                                                 fv = 1.340781e+154                                                 fv = 11.7095                                                 fv = 11.7095                                                 fv = 5.92879                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2.949878                                                 fv = 2.949878                                                 fv = 2.949878                                                 fv = 2.829853                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2.829853                                                 fv = 13.27787                                                 fv = 2.829853                                                 fv = 9.01141                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2.829853                                                 fv = 14.54335                                                 fv = 356.9508                                                 fv = 4.263958                                                 fv = 14.54335                                                 fv = 8.1049                                                 fv = 1.340781e+154                                                 fv = 3.233306                                                 fv = 3.2187e+13                                                 fv = 2.829853                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.01168                                                 fv = 40.04952                                                 fv = 2.829853                                                 fv = 2.829853                                                 fv = 11.58956                                                 fv = 4.344526                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2.829853                                                 fv = 1.340781e+154                                                 fv = 4.211063                                                 fv = 3.199477                                                 fv = 661.5582                                                 fv = 1.340781e+154                                                 fv = 6.186184                                                 fv = 3.39902                                                 fv = 3.140247                                                 fv = 4.04032                                                 fv = 250.3296                                                 fv = 29.8821                                                 fv = 1.340781e+154                                                 fv = 610.0253                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2.829853                                                 fv = 2.881773                                                 fv = 2.829853                                                 fv = 1.340781e+154                                                 fv = 3.625303                                                 fv = 5.759558                                                 fv = 2.829853                                                 fv = 1.340781e+154                                                 fv = 5.231139                                                 fv = 2.829853                                                 fv = 1.340781e+154                                                 fv = 14.01168                                                 fv = 2.829853                                                 fv = 4.957169                                                 fv = 2.829853                                                 fv = 2.829853                                                 fv = 20.29134                                                 fv = 2.829853                                                 fv = 2.829853                                                 fv = 2.829853 run # 2: f=2.829853277
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 3: f=1.340780793e+154
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 730.8899                                                 fv = 29.8821                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 26.81697                                                 fv = 1526.869                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.06892                                                 fv = 3177.731                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.861159                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 59282.77                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 159.0911                                                 fv = 341.1383                                                 fv = 7.861159                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 190.5081                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8091.485                                                 fv = 341.1383                                                 fv = 1.340781e+154                                                 fv = 10.23538                                                 fv = 1.340781e+154                                                 fv = 7.861159                                                 fv = 1.340781e+154                                                 fv = 24017.69                                                 fv = 1.340781e+154                                                 fv = 7.861159                                                 fv = 230751.1                                                 fv = 39.58142                                                 fv = 29.8821                                                 fv = 8.245792                                                 fv = 11.83246                                                 fv = 1355426                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 29.8821                                                 fv = 29.8821                                                 fv = 29.8821                                                 fv = 8.245792                                                 fv = 8.245792                                                 fv = 9.091895                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 20.5195                                                 fv = 38781585010                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.50982                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 13.16128                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.861159                                                 fv = 15.59402                                                 fv = 1.340781e+154                                                 fv = 192.2457                                                 fv = 7.861159                                                 fv = 7.861159                                                 fv = 1.340781e+154                                                 fv = 7.861159                                                 fv = 48.19449                                                 fv = 38971.86                                                 fv = 7.861159                                                 fv = 1801960                                                 fv = 16.19625                                                 fv = 7.861159                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.96792                                                 fv = 7.861159                                                 fv = 7.861159                                                 fv = 9.923217                                                 fv = 112.2583                                                 fv = 11.53962                                                 fv = 7.861159                                                 fv = 7.861159                                                 fv = 1.340781e+154                                                 fv = 20699.85                                                 fv = 217.7934                                                 fv = 349.0284                                                 fv = 7.861159                                                 fv = 41.44292                                                 fv = 7.861159                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 154.8118                                                 fv = 14.57903                                                 fv = 1.340781e+154                                                 fv = 12.85271                                                 fv = 7.861159                                                 fv = 7.861159                                                 fv = 36.44729                                                 fv = 14.96758                                                 fv = 33.36396                                                 fv = 7.861159                                                 fv = 7.861159                                                 fv = 14.70272                                                 fv = 28.86993                                                 fv = 1.340781e+154                                                 fv = 16.19625                                                 fv = 15.59402                                                 fv = 7.861159                                                 fv = 1.340781e+154                                                 fv = 16.51927                                                 fv = 7.861159                                                 fv = 30.1443                                                 fv = 10.12508 run # 4: f=7.861158874
#> 
#> bb eval : 804
#> best    : 1.510565197
#> worst   : 1.340780793e+154
#> solution: x = ( 0 0 0 0 1 0 2 0 0 2 1 1 1 1 1 1 1 1 1 4 ) f(x) = 1.510565197
#> 
#>                                                 fv = 1.510565                                                                       Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 39.73252