Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  0  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  3  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  1  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 26.63704                                                 fv = 85.53486                                                 fv = 78.03278                                                 fv = 4679.903                                                 fv = 4960.521                                                 fv = 7.619334                                                 fv = 125.9958                                                 fv = 1.340781e+154                                                 fv = 10.45604                                                 fv = 16.90357                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.183106                                                 fv = 6.183106                                                 fv = 6.183106                                                 fv = 13.76492                                                 fv = 1642.889                                                 fv = 260003.7                                                 fv = 6.63235                                                 fv = 3.555072                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.94266                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 850.8649                                                 fv = 1.340781e+154                                                 fv = 3.555072                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.27418                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.27418                                                 fv = 27.40528                                                 fv = 6.806206                                                 fv = 16.83793                                                 fv = 16.27418                                                 fv = 4.273329                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 24.75011                                                 fv = 1.340781e+154                                                 fv = 6.915725                                                 fv = 70.92355                                                 fv = 734.2266                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 21.97589                                                 fv = 1021.775                                                 fv = 1.340781e+154                                                 fv = 32.33731                                                 fv = 5.006194                                                 fv = 1.340781e+154                                                 fv = 59390.59                                                 fv = 1.340781e+154                                                 fv = 59390.59                                                 fv = 719.714                                                 fv = 15.2871                                                 fv = 873.6672                                                 fv = 1.340781e+154                                                 fv = 687.4937                                                 fv = 1.340781e+154                                                 fv = 9.961514                                                 fv = 141677.4                                                 fv = 216.5241                                                 fv = 42.54581                                                 fv = 13.64824                                                 fv = 3.555072                                                 fv = 1.340781e+154                                                 fv = 11.7442                                                 fv = 3.555072                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 37.73016                                                 fv = 3.555072                                                 fv = 45087018096                                                 fv = 11.7456                                                 fv = 1.340781e+154                                                 fv = 549.2032                                                 fv = 42452.18                                                 fv = 27.31794                                                 fv = 34.50472                                                 fv = 786403.9                                                 fv = 3013.213                                                 fv = 47.50211                                                 fv = 463.2781                                                 fv = 66.71407                                                 fv = 29.03346                                                 fv = 6.022668                                                 fv = 15.74687                                                 fv = 3.555072                                                 fv = 3.555072                                                 fv = 11.43264                                                 fv = 1.964941                                                 fv = 1.340781e+154                                                 fv = 9113.876                                                 fv = 3.904386                                                 fv = 3.369491                                                 fv = 18.77123                                                 fv = 62.04568                                                 fv = 26.98398                                                 fv = 3.402477                                                 fv = 1.964941                                                 fv = 1.964941                                                 fv = 2.244379                                                 fv = 4.637367                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3.049464                                                 fv = 44.8156                                                 fv = 8.725119                                                 fv = 34923.08                                                 fv = 31.29716                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 371.7641                                                 fv = 9.464616                                                 fv = 1.340781e+154                                                 fv = 1.964941                                                 fv = 7.785659                                                 fv = 7.649259                                                 fv = 1.964941                                                 fv = 4.601343                                                 fv = 1.964941                                                 fv = 3.047139                                                 fv = 27.0599 run # 0: f=1.96494118
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.768032                                                 fv = 29.07545                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.72909                                                 fv = 12.03729                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.64455                                                 fv = 133.2059                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 39.86463                                                 fv = 30624.68                                                 fv = 1581.662                                                 fv = 1.340781e+154                                                 fv = 7.100609                                                 fv = 3667.194                                                 fv = 1945031417                                                 fv = 4.353658                                                 fv = 294.5231                                                 fv = 10.64568                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6.653935                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.56542                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.472818                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.49714                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.185611                                                 fv = 4.239928                                                 fv = 35750924676                                                 fv = 1.340781e+154                                                 fv = 29.49516                                                 fv = 6.357273                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.213416                                                 fv = 4.185611                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.213416                                                 fv = 1.340781e+154                                                 fv = 7.213416                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.714216                                                 fv = 4.418127                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3.824481                                                 fv = 3.824481                                                 fv = 1.340781e+154                                                 fv = 24.98168                                                 fv = 219893892                                                 fv = 11.98708                                                 fv = 3.824481                                                 fv = 3.824481                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3.532498                                                 fv = 1.340781e+154                                                 fv = 12.62876                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.397231                                                 fv = 8.48702                                                 fv = 3.938687                                                 fv = 5.012288                                                 fv = 73.0932                                                 fv = 1.340781e+154                                                 fv = 3.532498                                                 fv = 3.532498                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.337583                                                 fv = 22373033261                                                 fv = 1.340781e+154                                                 fv = 11.01319                                                 fv = 5.01892                                                 fv = 4.637391                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.155177                                                 fv = 247.1445                                                 fv = 1.340781e+154                                                 fv = 30.00932                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.345                                                 fv = 186.6655                                                 fv = 7.679284                                                 fv = 1.340781e+154                                                 fv = 3.532498                                                 fv = 3.532498                                                 fv = 3.532498                                                 fv = 5.696322                                                 fv = 1.340781e+154                                                 fv = 3.532498                                                 fv = 1707.733                                                 fv = 21.84677                                                 fv = 24.6719                                                 fv = 1.340781e+154                                                 fv = 5.041843                                                 fv = 1.340781e+154                                                 fv = 64.00204                                                 fv = 483.3124                                                 fv = 3.78138                                                 fv = 1.340781e+154                                                 fv = 5.17581                                                 fv = 4.157434                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.411065                                                 fv = 3.532498                                                 fv = 3.532498                                                 fv = 3.532498                                                 fv = 3.532498                                                 fv = 11.65748                                                 fv = 66.35107                                                 fv = 9.742286                                                 fv = 11.65748                                                 fv = 5.762479                                                 fv = 27.55319                                                 fv = 1.340781e+154                                                 fv = 4.81291                                                 fv = 1.340781e+154                                                 fv = 3.200742                                                 fv = 26.94453                                                 fv = 1.340781e+154                                                 fv = 6.628451                                                 fv = 3.200742                                                 fv = 3.352128                                                 fv = 4.533681                                                 fv = 3.200742                                                 fv = 1.340781e+154                                                 fv = 3.200742                                                 fv = 4.62207                                                 fv = 3.200742                                                 fv = 4.863372                                                 fv = 5.012288                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.012288                                                 fv = 4.761016                                                 fv = 3.357336                                                 fv = 3.746959                                                 fv = 5.233104                                                 fv = 8.27774                                                 fv = 5.397231                                                 fv = 8.164903 run # 1: f=3.200741574
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.16678                                                 fv = 29.07545                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 29.07545                                                 fv = 14.16678                                                 fv = 14.16678                                                 fv = 10.54183                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 69157632                                                 fv = 2.261186e+13                                                 fv = 15.9972                                                 fv = 15.9972                                                 fv = 10.54183                                                 fv = 1.340781e+154                                                 fv = 7.761719                                                 fv = 6.323078                                                 fv = 13.87076                                                 fv = 230714.1                                                 fv = 1.340781e+154                                                 fv = 50324107                                                 fv = 1.340781e+154                                                 fv = 6.323078                                                 fv = 6.323078                                                 fv = 16.69793                                                 fv = 10.62716                                                 fv = 1.340781e+154                                                 fv = 9.317017                                                 fv = 9.542105                                                 fv = 1047.453                                                 fv = 25.65083                                                 fv = 1.340781e+154                                                 fv = 23.2138                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 113.6663                                                 fv = 13.14579                                                 fv = 13.14579                                                 fv = 1.340781e+154                                                 fv = 11.76712                                                 fv = 19.91071                                                 fv = 19.91071                                                 fv = 35335039478                                                 fv = 19.81582                                                 fv = 19.91071                                                 fv = 18.97715                                                 fv = 10.5718                                                 fv = 1.340781e+154                                                 fv = 1388864839                                                 fv = 19.81582                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 19.79102                                                 fv = 1.340781e+154                                                 fv = 13.14579                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.70031                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9644.467                                                 fv = 15.89481                                                 fv = 8.151199                                                 fv = 9.289898                                                 fv = 8.606112                                                 fv = 1.340781e+154                                                 fv = 7.737536                                                 fv = 1.340781e+154                                                 fv = 1666.633                                                 fv = 6.323078                                                 fv = 242.4549                                                 fv = 6.323078                                                 fv = 13.89211                                                 fv = 12.8553                                                 fv = 7.889983                                                 fv = 1.340781e+154                                                 fv = 74653721                                                 fv = 17.74796                                                 fv = 39446.27                                                 fv = 1.340781e+154                                                 fv = 13.14098                                                 fv = 146.2418                                                 fv = 15.86759                                                 fv = 6.323078                                                 fv = 27.54598                                                 fv = 12.33608                                                 fv = 6.323078                                                 fv = 25.27858                                                 fv = 22.68588                                                 fv = 1.340781e+154                                                 fv = 6.323078                                                 fv = 19.38675                                                 fv = 6.323078                                                 fv = 12.23036                                                 fv = 1.340781e+154                                                 fv = 17.50461                                                 fv = 12.57951                                                 fv = 7.868311                                                 fv = 9.890656                                                 fv = 10.54183                                                 fv = 1.340781e+154                                                 fv = 13.14098                                                 fv = 11.063                                                 fv = 1.340781e+154                                                 fv = 12.06048                                                 fv = 19.54041                                                 fv = 20.54148                                                 fv = 19.31931                                                 fv = 18.73398                                                 fv = 5.116383                                                 fv = 378.7735                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.116383                                                 fv = 13.19937                                                 fv = 5.116383                                                 fv = 5.116383                                                 fv = 5.116383                                                 fv = 6.395529                                                 fv = 6.203437                                                 fv = 4.108937                                                 fv = 26.24701                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.726948                                                 fv = 12.85361                                                 fv = 4.392214                                                 fv = 4.020621                                                 fv = 40.68021                                                 fv = 1.340781e+154 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 4.889445                                                 fv = 5.523602                                                 fv = 4.020621                                                 fv = 6.653502                                                 fv = 1.340781e+154                                                 fv = 5.105184                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 6.011351                                                 fv = 10.02038                                                 fv = 4.287329                                                 fv = 12.46271                                                 fv = 53.7596                                                 fv = 8.339448 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 5.250938                                                 fv = 4.477463                                                 fv = 4.020621                                                 fv = 4.020621                                                 fv = 5.464676 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 4.889445                                                 fv = 10.4514 run # 2: f=4.020620936
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 254.3032                                                 fv = 29.07545                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 19.34237                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2341704                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.247879                                                 fv = 2061.263                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 59.73601                                                 fv = 1.340781e+154                                                 fv = 7405053                                                 fv = 1.340781e+154                                                 fv = 45.07628                                                 fv = 960.5221                                                 fv = 1864.909                                                 fv = 4.247879                                                 fv = 5313.175                                                 fv = 10.02111                                                 fv = 102895.6                                                 fv = 1.340781e+154                                                 fv = 45.07628                                                 fv = 1.340781e+154                                                 fv = 3.755355                                                 fv = 1.340781e+154                                                 fv = 17.60106                                                 fv = 58191425                                                 fv = 3.131668                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 13.10789                                                 fv = 34.97299                                                 fv = 1.340781e+154                                                 fv = 4172.66                                                 fv = 56843163                                                 fv = 296.7399                                                 fv = 3.131668                                                 fv = 47601492                                                 fv = 1.340781e+154                                                 fv = 3.99438                                                 fv = 69484794200                                                 fv = 31.83598                                                 fv = 484931067                                                 fv = 406.8189                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.670847                                                 fv = 3.131668                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3.917131                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.974851                                                 fv = 5.697913                                                 fv = 123865037                                                 fv = 5.974851                                                 fv = 20747.8                                                 fv = 5144.881                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 36665478                                                 fv = 2343.814                                                 fv = 1.340781e+154                                                 fv = 11010.78                                                 fv = 114.9344                                                 fv = 1.340781e+154                                                 fv = 426.383                                                 fv = 1.340781e+154                                                 fv = 34.30471                                                 fv = 1.340781e+154                                                 fv = 4.812378                                                 fv = 19.32883                                                 fv = 3.131668                                                 fv = 64.83877                                                 fv = 3.131668                                                 fv = 3.131668                                                 fv = 1.936261                                                 fv = 2.413045                                                 fv = 1.340781e+154                                                 fv = 1963.729                                                 fv = 38.45611                                                 fv = 1.340781e+154                                                 fv = 416.9475                                                 fv = 145.4172                                                 fv = 11.45479                                                 fv = 11.95836                                                 fv = 94.70163                                                 fv = 1.936261                                                 fv = 698.3012                                                 fv = 1.936261                                                 fv = 94.70163                                                 fv = 1.936261                                                 fv = 1.340781e+154                                                 fv = 1.961896                                                 fv = 160.8397                                                 fv = 494.641                                                 fv = 81.9117                                                 fv = 1.340781e+154                                                 fv = 67520.86                                                 fv = 2.491425                                                 fv = 3.358086                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 29.2917                                                 fv = 3.753564                                                 fv = 5.458085                                                 fv = 5.163903                                                 fv = 40.58272                                                 fv = 10.65516                                                 fv = 28.37506                                                 fv = 904.2152                                                 fv = 1.936261                                                 fv = 4.734284                                                 fv = 1.340781e+154                                                 fv = 1.936261                                                 fv = 12.98678                                                 fv = 30.16343                                                 fv = 8.460536                                                 fv = 1.936261                                                 fv = 1.936261                                                 fv = 156.5893                                                 fv = 9.522604                                                 fv = 343.8634                                                 fv = 7.578689                                                 fv = 21.8086                                                 fv = 2.351355                                                 fv = 1.936261                                                 fv = 4.242395                                                 fv = 5.408165                                                 fv = 1.936261                                                 fv = 6.717014                                                 fv = 11.95836                                                 fv = 2.413045                                                 fv = 2.413045                                                 fv = 4.979641                                                 fv = 12.18012                                                 fv = 40.58272                                                 fv = 380.2393                                                 fv = 4.849611 run # 3: f=1.936261444
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 44.96169                                                 fv = 29.07545                                                 fv = 23.79166                                                 fv = 23.79166                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1865.64                                                 fv = 7.906875                                                 fv = 13.36361                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.684744                                                 fv = 7.906875                                                 fv = 7.906875                                                 fv = 461234640673                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 235012740684                                                 fv = 1.340781e+154                                                 fv = 9.995113                                                 fv = 7.906875                                                 fv = 1.340781e+154                                                 fv = 10.47057                                                 fv = 7.906875                                                 fv = 13.36361                                                 fv = 11.46501                                                 fv = 8.480133                                                 fv = 3.811158                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 204.6626                                                 fv = 1.340781e+154                                                 fv = 3.811158                                                 fv = 7.831655                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 53.8274                                                 fv = 3589.803                                                 fv = 625.6202                                                 fv = 109.9084                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.21425                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 37.62196                                                 fv = 3.811158                                                 fv = 8.482543                                                 fv = 3.811158                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 22.44683                                                 fv = 144701                                                 fv = 22.92028                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 540.8042                                                 fv = 196.5147                                                 fv = 158.2843                                                 fv = 8422.548                                                 fv = 3.811158                                                 fv = 3.811158                                                 fv = 3.811158                                                 fv = 3.811158                                                 fv = 3.811158                                                 fv = 1.340781e+154                                                 fv = 3.811158                                                 fv = 3.811158                                                 fv = 1.340781e+154                                                 fv = 22.92028                                                 fv = 15.58603                                                 fv = 5.214678                                                 fv = 828.625                                                 fv = 3.811158                                                 fv = 22.53563                                                 fv = 1335.583                                                 fv = 4.71499                                                 fv = 3.811158                                                 fv = 3.811158                                                 fv = 3.811158                                                 fv = 3.811158                                                 fv = 1.340781e+154                                                 fv = 3.811158                                                 fv = 3.811158                                                 fv = 3.811158                                                 fv = 5.164395                                                 fv = 20.07923 run # 4: f=3.811157985
#> 
#> bb eval : 821
#> best    : 1.936261444
#> worst   : 4.020620936
#> solution: x = ( 0 0 1 0 1 0 2 0 0 3 1 1 1 2 4 1 1 2 10 1 ) f(x) = 1.936261444
#> 
#>                                                 fv = 1.936261                                                                       Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 12.72911