Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  0  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  3  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  1  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 36.13523                                                 fv = 47.30838                                                 fv = 44.03764                                                 fv = 784.2506                                                 fv = 1169.444                                                 fv = 5.058302                                                 fv = 55.53303                                                 fv = 1.340781e+154                                                 fv = 8.120501                                                 fv = 22.95244                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.83499                                                 fv = 11.29791                                                 fv = 15.10918                                                 fv = 5.058302                                                 fv = 11.29791                                                 fv = 11.29791                                                 fv = 235.4809                                                 fv = 296.7936                                                 fv = 22.95244                                                 fv = 30.17004                                                 fv = 13.74752                                                 fv = 1.340781e+154                                                 fv = 29682.13                                                 fv = 11935654                                                 fv = 29.57303                                                 fv = 5.058302                                                 fv = 5.058302                                                 fv = 5.058302                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 28947.32                                                 fv = 7.760206                                                 fv = 1.340781e+154                                                 fv = 34.32734                                                 fv = 15.07629                                                 fv = 310.3336                                                 fv = 13560.99 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 6.313948                                                 fv = 1.340781e+154                                                 fv = 20444.89                                                 fv = 24.86588                                                 fv = 15.45671                                                 fv = 1.340781e+154                                                 fv = 26.98594                                                 fv = 8.418041                                                 fv = 5.058302                                                 fv = 33.34484 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 1.340781e+154                                                 fv = 5.058302                                                 fv = 5.058302                                                 fv = 12.69857                                                 fv = 1.340781e+154 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 1.340781e+154                                                 fv = 11.82726                                                 fv = 23.28504 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 15.72889 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 7.103676 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 6.818679 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 9.619073                                                 fv = 74.9067                                                 fv = 12.66422 run # 0: f=5.058301514
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4985.853                                                 fv = 40.72985                                                 fv = 40.72985                                                 fv = 40.72985                                                 fv = 27.38068                                                 fv = 27.38068                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 27.38068                                                 fv = 1.340781e+154                                                 fv = 29.01805                                                 fv = 10.63787                                                 fv = 1.340781e+154                                                 fv = 5317.414                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.63787                                                 fv = 1.340781e+154                                                 fv = 8.031307                                                 fv = 7.224178                                                 fv = 7.224178                                                 fv = 16.33236                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.224178                                                 fv = 1.340781e+154                                                 fv = 7.224178                                                 fv = 1.340781e+154                                                 fv = 7.295158                                                 fv = 7.295158                                                 fv = 93.93234                                                 fv = 10.22097                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.224178                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.43934                                                 fv = 2447925453                                                 fv = 9.760203                                                 fv = 1.340781e+154                                                 fv = 13.55589                                                 fv = 1.340781e+154                                                 fv = 17.02269                                                 fv = 36.41802                                                 fv = 7.180729                                                 fv = 7348157157                                                 fv = 1.340781e+154                                                 fv = 27.38068                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 34.32072                                                 fv = 1.340781e+154                                                 fv = 7.224178                                                 fv = 1.340781e+154                                                 fv = 7.180729                                                 fv = 8.446483                                                 fv = 7.180729                                                 fv = 1.340781e+154                                                 fv = 11.30242                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 3365.311                                                 fv = 1.340781e+154                                                 fv = 8057435                                                 fv = 15929.36                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.928382                                                 fv = 58.6694                                                 fv = 105.8793                                                 fv = 3034599                                                 fv = 18.41354                                                 fv = 51.31319                                                 fv = 113.4951                                                 fv = 21.86091                                                 fv = 21.86091                                                 fv = 623429524                                                 fv = 122.352                                                 fv = 354.4588                                                 fv = 7.180729                                                 fv = 1.340781e+154                                                 fv = 21.86091                                                 fv = 20.09092                                                 fv = 7.180729                                                 fv = 155.4246                                                 fv = 1.340781e+154                                                 fv = 4741.844                                                 fv = 7.180729                                                 fv = 1.340781e+154                                                 fv = 38915.29                                                 fv = 39.73887                                                 fv = 19.82721                                                 fv = 1.340781e+154                                                 fv = 9.841368                                                 fv = 19.40303                                                 fv = 30.8977                                                 fv = 15.12205                                                 fv = 10.07672                                                 fv = 1.340781e+154                                                 fv = 25.97637                                                 fv = 21.30467                                                 fv = 1.340781e+154                                                 fv = 7.180729                                                 fv = 7.180729                                                 fv = 13.63165                                                 fv = 7.180729                                                 fv = 37.07766                                                 fv = 155.819                                                 fv = 7.180729                                                 fv = 424.1568                                                 fv = 1.340781e+154                                                 fv = 32.41655                                                 fv = 5.626318                                                 fv = 44.9005                                                 fv = 10521.82                                                 fv = 1.340781e+154                                                 fv = 5.626318                                                 fv = 5.626318                                                 fv = 29.14227                                                 fv = 21.13174                                                 fv = 5.626318                                                 fv = 15.08306                                                 fv = 24.08329                                                 fv = 7.734741                                                 fv = 1.340781e+154                                                 fv = 5.626318                                                 fv = 5.626318                                                 fv = 7.337893                                                 fv = 20.83815                                                 fv = 16349.11                                                 fv = 15.84058                                                 fv = 8.685081                                                 fv = 100.2715                                                 fv = 3634.211                                                 fv = 1.340781e+154 run # 1: f=5.626317518
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 43.00118                                                 fv = 40.72985                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 20307354324                                                 fv = 40.72985                                                 fv = 40.72985                                                 fv = 40.72985                                                 fv = 8.47191                                                 fv = 385972158                                                 fv = 67959.35                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 781.6993                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 22.86605                                                 fv = 46.59823                                                 fv = 17.75697                                                 fv = 46.59823                                                 fv = 1.326768e+12                                                 fv = 1724926                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 128277379952                                                 fv = 781.6993                                                 fv = 1109591                                                 fv = 133.6112                                                 fv = 130.8222                                                 fv = 52.08377                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2829484980                                                 fv = 71.87537                                                 fv = 8.47191                                                 fv = 2.030438e+13                                                 fv = 10.42335                                                 fv = 8.47191                                                 fv = 10.65863                                                 fv = 1.340781e+154                                                 fv = 35.33644                                                 fv = 34.67628                                                 fv = 14.88494                                                 fv = 81311810993                                                 fv = 1.340781e+154                                                 fv = 2.23702e+12                                                 fv = 1.340781e+154                                                 fv = 135.6165                                                 fv = 72.10047                                                 fv = 14.88494                                                 fv = 8.47191                                                 fv = 14.88494                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 32.72897                                                 fv = 548.6839                                                 fv = 1.340781e+154                                                 fv = 8.961278                                                 fv = 8626823                                                 fv = 11.67246                                                 fv = 2027691                                                 fv = 1.340781e+154                                                 fv = 9.266537                                                 fv = 1944.587                                                 fv = 23.11552                                                 fv = 1.340781e+154                                                 fv = 301661.6                                                 fv = 8.47191                                                 fv = 8.47191                                                 fv = 8.47191                                                 fv = 130.9826                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.47191                                                 fv = 32.72897                                                 fv = 32.72897                                                 fv = 11.15423                                                 fv = 22.22675                                                 fv = 8.47191                                                 fv = 10.71986                                                 fv = 14.97457                                                 fv = 47.25732                                                 fv = 1.340781e+154                                                 fv = 11.44173                                                 fv = 1.340781e+154                                                 fv = 8.47191                                                 fv = 8.47191                                                 fv = 8.47191                                                 fv = 8.47191                                                 fv = 8.47191                                                 fv = 27.0434                                                 fv = 8.47191                                                 fv = 180.4603                                                 fv = 17.55442                                                 fv = 28.74473 run # 2: f=8.471909507
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 35.96517                                                 fv = 40.72985                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.600871                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.00032                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 13.12495                                                 fv = 133.1713                                                 fv = 8.877704                                                 fv = 6.29167                                                 fv = 1752761                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 82.15808                                                 fv = 82.15808                                                 fv = 25586.26                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4615.602                                                 fv = 213.8619                                                 fv = 17.49202                                                 fv = 1.340781e+154                                                 fv = 416.861                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 367.941                                                 fv = 219.316                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.63746                                                 fv = 100.0427                                                 fv = 6.29167                                                 fv = 8168136                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6379.672                                                 fv = 25586.26                                                 fv = 1.340781e+154                                                 fv = 1679.246                                                 fv = 1.340781e+154                                                 fv = 6.29167                                                 fv = 1.340781e+154                                                 fv = 405201.1                                                 fv = 12790.15                                                 fv = 2083.605                                                 fv = 12790.15                                                 fv = 25586.26                                                 fv = 25586.26                                                 fv = 27.38068                                                 fv = 1.340781e+154                                                 fv = 21.80654                                                 fv = 6047071                                                 fv = 7.228239                                                 fv = 8.40831                                                 fv = 137.7701                                                 fv = 12.26962                                                 fv = 59024.03                                                 fv = 1.340781e+154                                                 fv = 65.91896                                                 fv = 1.340781e+154                                                 fv = 6.29167                                                 fv = 543.0578                                                 fv = 6.29167                                                 fv = 10.30266                                                 fv = 54.24277                                                 fv = 6.29167                                                 fv = 3.03897                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.6484                                                 fv = 7.214139                                                 fv = 41469067                                                 fv = 1.340781e+154                                                 fv = 3.03897                                                 fv = 3.03897                                                 fv = 246.2641                                                 fv = 3.03897                                                 fv = 3.03897                                                 fv = 6.392863                                                 fv = 76.11121                                                 fv = 3.03897                                                 fv = 7.214139                                                 fv = 130.0369                                                 fv = 7.214139                                                 fv = 1.340781e+154                                                 fv = 7.214139                                                 fv = 194.7524                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 139939.7                                                 fv = 92.53891                                                 fv = 122.9145                                                 fv = 1.340781e+154                                                 fv = 3.03897                                                 fv = 1.328375                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 38.58289                                                 fv = 12.86994                                                 fv = 1.340781e+154                                                 fv = 1.328375                                                 fv = 1.340781e+154                                                 fv = 1.328375                                                 fv = 1.328375                                                 fv = 2.220852                                                 fv = 866.7767                                                 fv = 5.922658                                                 fv = 1.328375                                                 fv = 3.205454                                                 fv = 21.48734                                                 fv = 1.340781e+154                                                 fv = 197.8198                                                 fv = 1.328375                                                 fv = 1.876953                                                 fv = 1.340781e+154                                                 fv = 432.7567                                                 fv = 111.652                                                 fv = 1.340781e+154                                                 fv = 20.74538                                                 fv = 15.026                                                 fv = 5323.546                                                 fv = 1.328375                                                 fv = 20.86095                                                 fv = 30.77027                                                 fv = 11.2157                                                 fv = 1.328375                                                 fv = 454.5726                                                 fv = 1.340781e+154                                                 fv = 14.84652                                                 fv = 2.016149                                                 fv = 1.340781e+154                                                 fv = 1.328375                                                 fv = 3.215817                                                 fv = 10.09922                                                 fv = 1.743343                                                 fv = 1.328375                                                 fv = 6.329975                                                 fv = 1003.444                                                 fv = 1.328375 run # 3: f=1.328374713
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 40.72985                                                 fv = 40.72985                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.41395                                                 fv = 1065206                                                 fv = 83.90818                                                 fv = 1.340781e+154                                                 fv = 8.086191                                                 fv = 8.086191                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 13037.86                                                 fv = 4427893                                                 fv = 8206973                                                 fv = 3360.932                                                 fv = 1.340781e+154                                                 fv = 43.54779                                                 fv = 1.340781e+154                                                 fv = 21.06177                                                 fv = 6882058                                                 fv = 1.340781e+154                                                 fv = 43.54779                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15596450759                                                 fv = 8.086191                                                 fv = 8.086191                                                 fv = 59.666                                                 fv = 1.340781e+154                                                 fv = 10.41395                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 2402.66                                                 fv = 4.741427                                                 fv = 70.51587                                                 fv = 1.340781e+154                                                 fv = 13.24692                                                 fv = 4.741427                                                 fv = 5.526457                                                 fv = 9.184166                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.526457                                                 fv = 1.340781e+154                                                 fv = 23.20119                                                 fv = 728.6018                                                 fv = 18885.63                                                 fv = 5206043                                                 fv = 1.340781e+154                                                 fv = 4.741427                                                 fv = 3311.293                                                 fv = 4.741427                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 9.184166                                                 fv = 10.41395                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 145437.8                                                 fv = 2663.269                                                 fv = 10.41395                                                 fv = 4.741427                                                 fv = 6439.642                                                 fv = 20.63569                                                 fv = 50.28265                                                 fv = 4.741427                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.41395                                                 fv = 10.5385                                                 fv = 6.821755                                                 fv = 1.340781e+154                                                 fv = 89.42423                                                 fv = 191.7322                                                 fv = 5.820994                                                 fv = 4.741427                                                 fv = 4.741427                                                 fv = 495.8438                                                 fv = 15.67033                                                 fv = 1.340781e+154                                                 fv = 7.206959                                                 fv = 4.741427                                                 fv = 25.08767                                                 fv = 4.741427                                                 fv = 4.741427                                                 fv = 1.340781e+154                                                 fv = 14.95716                                                 fv = 1.340781e+154                                                 fv = 5.526457                                                 fv = 1.340781e+154                                                 fv = 7.924609                                                 fv = 12.33675                                                 fv = 4.914018                                                 fv = 4.741427                                                 fv = 24.48986                                                 fv = 7.092607                                                 fv = 4.741427                                                 fv = 4.741427                                                 fv = 4.741427                                                 fv = 46.11727 run # 4: f=4.741427236
#> 
#> bb eval : 641
#> best    : 1.328374713
#> worst   : 8.471909507
#> solution: x = ( 1 4 0 4 0 0 4 0 0 1 1 1 1 1 4 1 2 2 4 2 ) f(x) = 1.328374713
#> 
#>                                                 fv = 1.328375                                                                       
#> Warning:  optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (4): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max
#> Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 232.5511