Regression Splines Learner

Categorical Regression Splines. Calls crs::crs() from crs.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.crs")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, crs

Parameters

Id	Type	Default	Levels	Range

degree	integer	3		\([0, \infty)\)

segments	integer	1		\([1, \infty)\)

include	integer	-		\((-\infty, \infty)\)

lambda	untyped	-		-

lambda.discrete	logical	FALSE	TRUE, FALSE	-

lambda.discrete.num	integer	100		\([0, \infty)\)

cv	character	nomad	nomad, exhaustive, none	-

cv.threshold	integer	10000		\([0, \infty)\)

cv.func	character	cv.ls	cv.ls, cv.gcv, cv.aic	-

kernel	logical	TRUE	TRUE, FALSE	-

degree.max	integer	10		\([0, \infty)\)

segments.max	integer	10		\([1, \infty)\)

degree.min	integer	0		\([0, \infty)\)

segments.min	integer	1		\([1, \infty)\)

cv.df.min	integer	1		\((-\infty, \infty)\)

complexity	character	degree-knots	degree-knots, degree, knots	-

knots	character	quantiles	quantiles, uniform, auto	-

basis	character	auto	auto, additive, tensor, glp	-

prune	logical	FALSE	TRUE, FALSE	-

restarts	integer	0		\([0, \infty)\)

nmulti	integer	5		\([0, \infty)\)

singular.ok	logical	FALSE	TRUE, FALSE	-

deriv	integer	0		\([0, \infty)\)

data.return	logical	FALSE	TRUE, FALSE	-

model.return	logical	FALSE	TRUE, FALSE	-

random.seed	integer	-		\((-\infty, \infty)\)

tau	numeric	-		\([0, 1]\)

initial.mesh.size.real	untyped	-		-

initial.mesh.size.integer	untyped	-		-

max.bb.eval	untyped	-		-

min.mesh.size.real	untyped	-		-

min.mesh.size.integer	untyped	-		-

min.poll.size.real	untyped	-		-

min.poll.size.integer	untyped	-		-

opts	untyped	-		-

See also

Author

annanzrv

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrCrs

Methods

Public methods

LearnerRegrCrs$new()

LearnerRegrCrs$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrCrs$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrCrs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.crs")
print(learner)
#> 
#> ── <LearnerRegrCrs> (regr.crs): Regression Splines ─────────────────────────────
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3 and crs
#> • Predict Types: [response] and se
#> • Feature Types: integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> Calling NOMAD (Nonsmooth Optimization by Mesh Adaptive Direct Search)
#> 
#> starting point # 0: (  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 )
#> starting point # 1: (  0  0  1  7  5  1  6  0  0  1  4  5  8  7  1  3  6  6  1  1 )
#> starting point # 2: (  0  0  1  9  8  1  1  3  0  3  8  6  3  6  4  5  2  6  7  7 )
#> starting point # 3: (  0  3  0  5  1  0  8  8  0  6  3  2  6  2  9  9  8  2  9  9 )
#> starting point # 4: (  0  1  0  0  6  1  4  5  0  9  7  9  4  3  6  7  5  9  4  3 )
#> 
#>                                                 fv = 23.69271                                                 fv = 28.32277                                                 fv = 23.06782                                                 fv = 797.3699                                                 fv = 1.340781e+154                                                 fv = 34.14892                                                 fv = 49.86688                                                 fv = 467753.5                                                 fv = 1.340781e+154                                                 fv = 40.08537                                                 fv = 47.66432                                                 fv = 1.340781e+154                                                 fv = 497.5647                                                 fv = 1.340781e+154                                                 fv = 7.894974                                                 fv = 1.340781e+154                                                 fv = 149.2225                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 161.1287                                                 fv = 46.42772                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 35.07515                                                 fv = 4049.342                                                 fv = 7256.782                                                 fv = 1.340781e+154                                                 fv = 15.0461                                                 fv = 1.340781e+154                                                 fv = 46.42772                                                 fv = 651.4856                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.894974                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 53.09693                                                 fv = 8.470141                                                 fv = 269281.9                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 60.15167                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 13.34356                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.89991                                                 fv = 17.45722                                                 fv = 16.65805                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.57332                                                 fv = 18554.23                                                 fv = 202.985                                                 fv = 1.340781e+154                                                 fv = 25.92364                                                 fv = 3291.293                                                 fv = 214.4756                                                 fv = 1.465078                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 1.340781e+154                                                 fv = 5.518585                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.44511                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 70.91258                                                 fv = 81.50154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.465078                                                 fv = 1.340781e+154                                                 fv = 0.5153078                                                 fv = 1.340781e+154 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 1.340781e+154                                                 fv = 24.24479                                                 fv = 1.340781e+154                                                 fv = 24.94018                                                 fv = 28.92291                                                 fv = 64.72572                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 29.74257                                                 fv = 1.340781e+154                                                 fv = 0.5153078                                                 fv = 19.24557                                                 fv = 0.5153078                                                 fv = 4.771071                                                 fv = 6.278486                                                 fv = 1.340781e+154                                                 fv = 0.5153078                                                 fv = 4.617552                                                 fv = 0.3972594                                                 fv = 1.340781e+154                                                 fv = 6.309599                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 46.27949                                                 fv = 1.487614                                                 fv = 1.487614                                                 fv = 10464.94                                                 fv = 11.71673                                                 fv = 9.750243                                                 fv = 0.3972594                                                 fv = 0.3972594                                                 fv = 1.340781e+154                                                 fv = 0.5086348                                                 fv = 0.3650251                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.94836                                                 fv = 6.405392                                                 fv = 56.86857                                                 fv = 66.15951                                                 fv = 260.1711                                                 fv = 17.84825                                                 fv = 3.538975                                                 fv = 0.3650251                                                 fv = 0.3650251                                                 fv = 1.340781e+154                                                 fv = 0.3650251                                                 fv = 0.3650251                                                 fv = 2119.081                                                 fv = 19.75026                                                 fv = 29.56379                                                 fv = 218.0849                                                 fv = 97.31582                                                 fv = 12.24673                                                 fv = 260.1711                                                 fv = 149.1184                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.857835                                                 fv = 10.09574                                                 fv = 0.7931218                                                 fv = 3.873343                                                 fv = 39.57243                                                 fv = 19.26264                                                 fv = 17.95959                                                 fv = 1.340781e+154                                                 fv = 20376.13                                                 fv = 8577.29                                                 fv = 0.2852111                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.03769                                                 fv = 0.2852111                                                 fv = 0.2852111                                                 fv = 0.2852111                                                 fv = 212.9222                                                 fv = 0.2852111                                                 fv = 1718.565                                                 fv = 131.574                                                 fv = 1.340781e+154                                                 fv = 15.55626                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 0.2852111                                                 fv = 524.2587                                                 fv = 38.53786                                                 fv = 7.145491                                                 fv = 19.75026                                                 fv = 1.479245                                                 fv = 1.340781e+154 run # 0: f=0.2852110546
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 run # 1: f=1.340780793e+154
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 116.968                                                 fv = 41.28765                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 19.25387                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 31865043                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 81.77705                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 19.25387                                                 fv = 14.54981                                                 fv = 14.54981                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 14.54981                                                 fv = 50.67431                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.41234                                                 fv = 23.54179                                                 fv = 132.2828                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 348.5449                                                 fv = 18.7663                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.6226                                                 fv = 1.340781e+154                                                 fv = 10.41234                                                 fv = 1.340781e+154                                                 fv = 671300.8                                                 fv = 1.340781e+154                                                 fv = 10.41234                                                 fv = 10.41234                                                 fv = 1.340781e+154                                                 fv = 1128.837                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1431.649                                                 fv = 10.41234                                                 fv = 10.41234                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 712999497                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.41234                                                 fv = 10.41234                                                 fv = 12.04788                                                 fv = 10.41234                                                 fv = 8.429718                                                 fv = 351.9932                                                 fv = 1.340781e+154                                                 fv = 15.62062                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.697357                                                 fv = 5.697357                                                 fv = 1.340781e+154                                                 fv = 41.28765                                                 fv = 11.5102                                                 fv = 8.916633                                                 fv = 16.16196                                                 fv = 1.340781e+154                                                 fv = 7.454759                                                 fv = 11.26551                                                 fv = 5.697357                                                 fv = 5.697357                                                 fv = 1.340781e+154                                                 fv = 7.454759                                                 fv = 1.340781e+154                                                 fv = 5.697357                                                 fv = 6.548076                                                 fv = 1.340781e+154                                                 fv = 8.916633                                                 fv = 1.340781e+154                                                 fv = 26.38564                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.0103                                                 fv = 6.063926                                                 fv = 10.41436                                                 fv = 8.916633                                                 fv = 9.989622                                                 fv = 7.037225                                                 fv = 7.797678                                                 fv = 7.037225                                                 fv = 5.697357                                                 fv = 5.697357                                                 fv = 5.697357                                                 fv = 5.697357                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.255265                                                 fv = 6.701968                                                 fv = 5.697357                                                 fv = 5.697357                                                 fv = 6.246569                                                 fv = 1428.807                                                 fv = 41.28765                                                 fv = 26.76299                                                 fv = 24.48117                                                 fv = 1.340781e+154                                                 fv = 6.540504                                                 fv = 7.385767                                                 fv = 6.548076                                                 fv = 5.6056                                                 fv = 3183.536                                                 fv = 1.340781e+154                                                 fv = 10.55923                                                 fv = 5.6056                                                 fv = 7.391883                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5.6056                                                 fv = 6.133368                                                 fv = 5.6056                                                 fv = 5.6056                                                 fv = 5.6056                                                 fv = 6.614665                                                 fv = 5.697357                                                 fv = 5.697357                                                 fv = 1.340781e+154                                                 fv = 5.697357                                                 fv = 6.426579                                                 fv = 6.520212                                                 fv = 6.727183                                                 fv = 7.758121                                                 fv = 8.837192                                                 fv = 2400.035 run # 2: f=5.605599951
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 110.4646                                                 fv = 41.28765                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1684364997                                                 fv = 41.28765                                                 fv = 16.61474                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.75001                                                 fv = 8758.399                                                 fv = 572990.5                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.75001                                                 fv = 15.75001                                                 fv = 15.75001                                                 fv = 17.51449                                                 fv = 15.75001                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 23.1949                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7118004                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 6726885                                                 fv = 1.340781e+154                                                 fv = 14.4383                                                 fv = 14.4383                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.84827                                                 fv = 9.036851                                                 fv = 26876.53                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 11.30324                                                 fv = 211.0619                                                 fv = 7.137097                                                 fv = 14.4383                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.137097                                                 fv = 1.340781e+154                                                 fv = 9.66219                                                 fv = 9.912736                                                 fv = 10.37547                                                 fv = 23.7419                                                 fv = 60.69847                                                 fv = 21.1785                                                 fv = 22.07124                                                 fv = 109.5337                                                 fv = 60.46278                                                 fv = 1169.104                                                 fv = 347.4646                                                 fv = 6013.307                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 10.1411                                                 fv = 40899915                                                 fv = 1.340781e+154                                                 fv = 1386.9                                                 fv = 13724.98                                                 fv = 19.67769                                                 fv = 12.43258                                                 fv = 34.8571                                                 fv = 7.991087                                                 fv = 9.036851                                                 fv = 32.1697                                                 fv = 18.3853                                                 fv = 1.340781e+154                                                 fv = 9.162788                                                 fv = 9.162788                                                 fv = 12.89733                                                 fv = 12.03745                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 136215.4                                                 fv = 323715.2                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 15.35574                                                 fv = 2.999943e+15                                                 fv = 148.8681                                                 fv = 1.340781e+154                                                 fv = 5.908793                                                 fv = 446.1205                                                 fv = 1.340781e+154                                                 fv = 23.1182                                                 fv = 10.396                                                 fv = 1.340781e+154                                                 fv = 6.130771                                                 fv = 1.340781e+154                                                 fv = 6.359393                                                 fv = 8.178237                                                 fv = 5.908793                                                 fv = 21.52997                                                 fv = 22.2118                                                 fv = 19.86503                                                 fv = 8.262167                                                 fv = 4.85365                                                 fv = 6.440342                                                 fv = 1.340781e+154                                                 fv = 9.407747                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 4.803074                                                 fv = 4.803074                                                 fv = 1.340781e+154                                                 fv = 16.58754                                                 fv = 3.289674                                                 fv = 1.340781e+154                                                 fv = 12422.46                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 28.99962                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 47.861                                                 fv = 1.340781e+154                                                 fv = 5.211879                                                 fv = 5.65068                                                 fv = 3204905                                                 fv = 14.1824                                                 fv = 3.289674                                                 fv = 41.68372                                                 fv = 1.340781e+154                                                 fv = 7.477415                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.206504                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 5959.73                                                 fv = 14198.13                                                 fv = 1725.657                                                 fv = 3.289674                                                 fv = 9.440256                                                 fv = 10.78931                                                 fv = 54.01545                                                 fv = 6.217462                                                 fv = 10.53825                                                 fv = 1285622                                                 fv = 4.84613                                                 fv = 1.340781e+154                                                 fv = 5.845832                                                 fv = 6.417228                                                 fv = 3.289674                                                 fv = 6.752369                                                 fv = 98.05596                                                 fv = 1.340781e+154                                                 fv = 75.01757                                                 fv = 3.289674                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 69.37824                                                 fv = 1.340781e+154                                                 fv = 24.1041                                                 fv = 3.493227                                                 fv = 1.340781e+154                                                 fv = 3.289674                                                 fv = 4.460925                                                 fv = 450.3469                                                 fv = 3.289674                                                 fv = 1.340781e+154                                                 fv = 6.017837                                                 fv = 436.7903                                                 fv = 1.340781e+154                                                 fv = 864.3795                                                 fv = 169.2223                                                 fv = 42.53069                                                 fv = 2.862064                                                 fv = 5.575304                                                 fv = 1.340781e+154                                                 fv = 69.37824                                                 fv = 5.096696                                                 fv = 29.08561                                                 fv = 6.18275                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 17.00102                                                 fv = 8.790029                                                 fv = 6.793981                                                 fv = 12.94271                                                 fv = 62.38368                                                 fv = 6.18275                                                 fv = 12.33529                                                 fv = 6.18275                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 16.91192                                                 fv = 6.939927                                                 fv = 20.57985                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 18225.39                                                 fv = 1.340781e+154                                                 fv = 11.02268                                                 fv = 6151.6                                                 fv = 14.77019                                                 fv = 4.268957                                                 fv = 1.340781e+154                                                 fv = 4.14863                                                 fv = 1.340781e+154                                                 fv = 55.66219                                                 fv = 220.5094                                                 fv = 2.862064                                                 fv = 4.268957                                                 fv = 6.34521                                                 fv = 5.974752                                                 fv = 5.575304                                                 fv = 2.862064                                                 fv = 2.862064                                                 fv = 2.862064                                                 fv = 10.0694                                                 fv = 6.957818 run # 3: f=2.862063825
#>                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.12391                                                 fv = 41.28765                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 7.12391                                                 fv = 11.71133                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                                                 fv = 1.340781e+154                                                 fv = 16.28146                                                 fv = 1.340781e+154                                                 fv = 16.3092                                                 fv = 27.54694                                                 fv = 25.34269                                                 fv = 7.12391                                                 fv = 348.1456                                                 fv = 21.59029                                                 fv = 18.2236                                                 fv = 1.340781e+154                                                 fv = 26.04613                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 12.16976                                                 fv = 12.06934                                                 fv = 18.84512                                                 fv = 1.340781e+154                                                 fv = 8.205834                                                 fv = 7.12391                                                 fv = 16.62349                                                 fv = 8.286213                                                 fv = 76.913                                                 fv = 7.12391                                                 fv = 1632.074                                                 fv = 236799428                                                 fv = 7.12391                                                 fv = 11.29169                                                 fv = 7.12391                                                 fv = 7.12391                                                 fv = 1.340781e+154                                                 fv = 136.6305                                                 fv = 1.340781e+154                                                 fv = 10.27153                                                 fv = 12.77149                                                 fv = 31.24752                                                 fv = 1.340781e+154                                                 fv = 1.340781e+154                                                 fv = 8.637837                                                 fv = 7.270507                                                 fv = 8.690807                                                 fv = 1.340781e+154                                                 fv = 17.5997                                                 fv = 7.12391                                                 fv = 8.572791                                                 fv = 7.12391                                                 fv = 7.12391                                                 fv = 297.6823                                                 fv = 62.35072                                                 fv = 12.76538                                                 fv = 7.827344                                                 fv = 7.12391                                                 fv = 10.8532                                                 fv = 7.12391                                                 fv = 7.12391                                                 fv = 10.98676                                                 fv = 19.4557                                                 fv = 178.0753 run # 4: f=7.123910399
#> 
#> bb eval : 763
#> best    : 0.2852110546
#> worst   : 1.340780793e+154
#> solution: x = ( 0 0 2 6 1 0 2 0 0 2 1 1 1 1 2 1 2 2 1 4 ) f(x) = 0.2852110546
#> 
#>                                                 fv = 0.2852111                                                                       
#> Warning:  optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (4): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (2): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max optimal degree equals search maximum (1): rerun with larger degree.max optimal degree equals search maximum (10): rerun with larger degree.max
#> Working...          

print(learner$model)
#> Call:
#> crs.formula(formula = formula, data = data, weights = private$.get_weights(task))

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Working...
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#> Warning: some 'x' values beyond boundary knots may cause ill-conditioned bases
#>           

# Score the predictions
predictions$score()
#> regr.mse 
#> 650.3909