Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 37.99761 35.98567 31.24158 27.21032 23.93915 21.31321 19.23369 17.61565
#>   [9] 16.38613 15.48254 14.85126 14.44640 14.22875 14.16486 14.22626 14.38552
#>  [17] 14.45378 14.42051 14.57140 14.74677 14.88543 14.91165 14.88225 14.68899
#>  [25] 14.74295 15.22298 15.57903 16.21282 16.29958 16.47469 16.66501 17.30092
#>  [33] 17.64830 17.55617 17.41625 16.93659 17.04611 16.91160 16.76274 15.78552
#>  [41] 15.73980 15.64379 17.60363 17.61309 17.61308 17.68243 21.22741 22.15988
#>  [49] 22.33308 23.19631 24.52841 26.79087 27.23194 29.42677 29.35755 31.74191
#>  [57] 31.74422 32.27356 29.58947 29.58872 29.58773 29.60866 32.87796 32.87805
#>  [65] 32.01898 32.01889 31.91852 31.91862 31.78324 31.47254 31.47327 31.47396
#>  [73] 31.47461 29.05456 29.05425 29.05394 29.05361 29.05328 29.05294 29.05259
#>  [81] 28.87725 28.85077 28.85006 28.84938 28.84871 28.84805 28.84741 28.84679
#>  [89] 28.84618 28.84559 28.84501 28.84444 28.84389 28.84336 28.84283 28.84232
#>  [97] 28.84182 28.84155 28.86730 28.86694
#> 
#> $cvse
#>   [1] 11.396684 11.252848 10.082984  9.023332  8.144830  7.407637  6.780931
#>   [8]  6.242007  5.775107  5.370142  5.021415  4.726391  4.484528  4.296162
#>  [15]  4.161475  4.079957  4.020416  3.948881  3.962018  3.970556  3.982341
#>  [22]  3.972027  3.960413  3.918818  3.930321  4.081003  4.231827  4.237027
#>  [29]  4.229727  4.225064  4.229998  4.211276  4.169799  4.154000  4.152141
#>  [36]  3.948981  3.968561  3.929465  3.949575  4.111896  4.088670  4.089221
#>  [43]  4.292921  4.293469  4.293546  4.353826  6.278488  6.073483  6.080952
#>  [50]  6.065979  5.929190  6.937634  6.882381  8.545779  8.556329  8.801179
#>  [57]  8.799930  9.328388  7.734569  7.734283  7.734036  7.746327  8.412884
#>  [64]  8.412903  8.458770  8.458796  8.456297  8.456306  8.468527  8.448966
#>  [71]  8.448989  8.449009  8.449028  7.782471  7.782242  7.782017  7.781797
#>  [78]  7.781582  7.781372  7.781166  7.769869  7.771792  7.771596  7.771406
#>  [85]  7.771220  7.771040  7.770865  7.770695  7.770529  7.770368  7.770212
#>  [92]  7.770060  7.769912  7.769768  7.769629  7.769493  7.769362  7.769218
#>  [99]  7.764225  7.764081
#> 
#> $fold
#>  [1]  6  1  5  3  1  8  2 10  9  2  3 10  5  4  7  8  9  7  4  1  6
#> 
#> $lambda
#>   [1] 5.114955180 4.770224320 4.448727166 4.148897844 3.869276015 3.608499762
#>   [7] 3.365298955 3.138489069 2.926965410 2.729697737 2.545725246 2.374151885
#>  [13] 2.214141994 2.064916234 1.925747791 1.795958835 1.674917220 1.562033405
#>  [19] 1.456757582 1.358576996 1.267013454 1.181620987 1.101983688 1.027713675
#>  [25] 0.958449213 0.893852944 0.833610247 0.777427707 0.725031681 0.676166972
#>  [31] 0.630595582 0.588095550 0.548459878 0.511495518 0.477022432 0.444872717
#>  [37] 0.414889784 0.386927602 0.360849976 0.336529896 0.313848907 0.292696541
#>  [43] 0.272969774 0.254572525 0.237415188 0.221414199 0.206491623 0.192574779
#>  [49] 0.179595883 0.167491722 0.156203341 0.145675758 0.135857700 0.126701345
#>  [55] 0.118162099 0.110198369 0.102771368 0.095844921 0.089385294 0.083361025
#>  [61] 0.077742771 0.072503168 0.067616697 0.063059558 0.058809555 0.054845988
#>  [67] 0.051149552 0.047702243 0.044487272 0.041488978 0.038692760 0.036084998
#>  [73] 0.033652990 0.031384891 0.029269654 0.027296977 0.025457252 0.023741519
#>  [79] 0.022141420 0.020649162 0.019257478 0.017959588 0.016749172 0.015620334
#>  [85] 0.014567576 0.013585770 0.012670135 0.011816210 0.011019837 0.010277137
#>  [91] 0.009584492 0.008938529 0.008336102 0.007774277 0.007250317 0.006761670
#>  [97] 0.006305956 0.005880955 0.005484599 0.005114955
#> 
#> $fit
#> $beta
#>               5.1150     4.7702     4.4487    4.1489    3.8693   3.6085
#> (Intercept) 21.03333 22.6436881 24.1455104 25.546115 26.852323 28.07050
#> am           0.00000  0.0000000  0.0000000  0.000000  0.000000  0.00000
#> carb         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.00000
#> cyl          0.00000  0.0000000  0.0000000  0.000000  0.000000  0.00000
#> disp         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.00000
#> drat         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.00000
#> gear         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.00000
#> hp           0.00000  0.0000000  0.0000000  0.000000  0.000000  0.00000
#> qsec         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.00000
#> vs           0.00000  0.0000000  0.0000000  0.000000  0.000000  0.00000
#> wt           0.00000 -0.5154549 -0.9961699 -1.444486 -1.862588 -2.25251
#>                3.3653    3.1385    2.9270    2.7297    2.5457    2.3742
#> (Intercept) 29.206571 30.266077 31.254176 32.175680 33.035078 33.836555
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.616154 -2.955289 -3.271567 -3.566529 -3.841612 -4.098155
#>                2.2141    2.0649    1.9257    1.7960    1.6749    1.5620
#> (Intercept) 34.584016 35.281100 35.931203 36.537491 36.962461 36.962461
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -4.337408 -4.560536 -4.768626 -4.962692 -5.098719 -5.098719
#>                1.4568      1.3586      1.2670     1.1816     1.1020    1.0277
#> (Intercept) 36.962461 36.68243274 35.27476755 33.9619016 32.7375185 31.595655
#> am           0.000000  0.00000000  0.00000000  0.0000000  0.0000000  0.000000
#> carb         0.000000  0.00000000  0.00000000  0.0000000  0.0000000  0.000000
#> cyl          0.000000  0.00000000  0.00000000  0.0000000  0.0000000  0.000000
#> disp         0.000000  0.00000000  0.00000000  0.0000000  0.0000000  0.000000
#> drat         0.000000  0.00000000  0.00000000  0.0000000  0.0000000  0.000000
#> gear         0.000000  0.00000000  0.00000000  0.0000000  0.0000000  0.000000
#> hp           0.000000  0.00000000  0.00000000  0.0000000  0.0000000  0.000000
#> qsec         0.000000  0.01467645  0.08845315  0.1572614  0.2214321  0.281278
#> vs           0.000000  0.00000000  0.00000000  0.0000000  0.0000000  0.000000
#> wt          -5.098719 -5.09409491 -5.07084913 -5.0491688 -5.0289497 -5.010093
#>                 0.9584     0.8939     0.8336     0.7774     0.7250    0.6762
#> (Intercept) 30.5307486 29.5376137 28.6114128 27.7476346 26.9420723 26.190802
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> qsec         0.3370904  0.3891413  0.4376841  0.4829554  0.5251754  0.564550
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> wt          -4.9925078 -4.9761074 -4.9608124 -4.9465482 -4.9332454 -4.920839
#>                 0.6306     0.5881     0.5485     0.5115    0.4770     0.4449
#> (Intercept) 25.4901651 24.8367486 24.2273702 23.6590619 10.734840 10.8207438
#> am           0.0000000  0.0000000  0.0000000  0.0000000  3.182819  3.1948824
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.000000 -0.0120747
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> qsec         0.6012709  0.6355169  0.6674548  0.6972403  1.122349  1.1169842
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> wt          -4.9092690 -4.8984787 -4.8884156 -4.8790307 -3.689602 -3.6775561
#>                 0.4149     0.3869     0.3608     0.3365     0.3138     0.2927
#> (Intercept) 12.0266604 13.1521204 14.2007463 14.9141593 14.9114610 14.9095058
#> am           3.3430431  3.4818119  3.6107304  3.6972821  3.6980726  3.6986373
#> carb        -0.1740257 -0.3253566 -0.4662143 -0.5616351 -0.5616880 -0.5617233
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         1.0432924  0.9744796  0.9103933  0.8668818  0.8669609  0.8670189
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -3.5210634 -3.3747109 -3.2385799 -3.1466511 -3.1463213 -3.1460873
#>                 0.2730     0.2546     0.2374     0.2214     0.2065     0.1926
#> (Intercept) 14.9080977 14.9070881 14.9063665 14.9058518 14.9054854 14.9052249
#> am           3.6990399  3.6993265  3.6995302  3.6996749  3.6997776  3.6998505
#> carb        -0.5617471 -0.5617633 -0.5617745 -0.5617823 -0.5617877 -0.5617915
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.8670610  0.8670913  0.8671131  0.8671286  0.8671398  0.8671477
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -3.1459214 -3.1458037 -3.1457202 -3.1456611 -3.1456192 -3.1455895
#>                  0.1796      0.1675      0.1562      0.1457      0.1359
#> (Intercept) 16.19125972 16.18942997 16.18793490 15.38807517 14.37143634
#> am           4.47261504  4.47232517  4.47207569  4.31968531  4.12827098
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.01499223  0.01499349  0.01499448  0.01564766  0.01651393
#> drat         0.00000000  0.00000000  0.00000000  0.27809544  0.62954305
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.03082415 -0.03081773 -0.03081235 -0.03108705 -0.03145078
#> qsec         0.84351767  0.84364258  0.84374539  0.83431120  0.82287401
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -3.63290123 -3.63337196 -3.63376279 -3.65529762 -3.68565585
#>                  0.1267      0.1182      0.1102      0.1028      0.0958
#> (Intercept) 12.14442935 12.14300522 12.14163339 11.78251326 13.36346737
#> am           4.85540359  4.85551492  4.85562216  4.77877307  5.21768734
#> carb         0.87747948  0.87764341  0.87780132  0.86659405  1.48419778
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.03214490  0.03214956  0.03215404  0.03227699  0.03802541
#> drat         0.00000000  0.00000000  0.00000000  0.13590301  1.52987450
#> gear         0.00000000  0.00000000  0.00000000  0.00000000 -1.89438795
#> hp          -0.05505555 -0.05505955 -0.05506341 -0.05487506 -0.05973366
#> qsec         1.13975485  1.13985140  1.13994441  1.13254100  1.28332462
#> vs           0.00000000  0.00000000  0.00000000  0.00000000 -0.82586630
#> wt          -4.92175566 -4.92214920 -4.92252828 -4.91847643 -6.19310055
#>                  0.0894      0.0834      0.0777      0.0725      0.0676
#> (Intercept) 13.36392503 13.36437429 13.36481490 13.36524669 13.36566955
#> am           5.21746871  5.21725544  5.21704744  5.21684463  5.21664690
#> carb         1.48396848  1.48374467  1.48352628  1.48331320  1.48310534
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.03802067  0.03801605  0.03801156  0.03800718  0.03800292
#> drat         1.52979391  1.52971351  1.52963351  1.52955408  1.52947536
#> gear        -1.89415243 -1.89392035 -1.89369192 -1.89346734 -1.89324673
#> hp          -0.05972904 -0.05972456 -0.05972022 -0.05971600 -0.05971191
#> qsec         1.28324532  1.28316755  1.28309133  1.28301668  1.28294361
#> vs          -0.82589853 -0.82592810 -0.82595530 -0.82598037 -0.82600354
#> wt          -6.19263859 -6.19218677 -6.19174506 -6.19131339 -6.19089168
#>                  0.0631      0.0588      0.0548      0.0511      0.0477
#> (Intercept) 13.36608340 13.36648822 13.36688402 13.36727084 13.36764875
#> am           5.21645414  5.21626624  5.21608311  5.21590463  5.21573070
#> carb         1.48290259  1.48270485  1.48251202  1.48232399  1.48214067
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.03799877  0.03799473  0.03799079  0.03798696  0.03798323
#> drat         1.52939748  1.52932056  1.52924469  1.52916995  1.52909641
#> gear        -1.89303022 -1.89281788 -1.89260979 -1.89240598 -1.89220648
#> hp          -0.05970794 -0.05970408 -0.05970033 -0.05969668 -0.05969314
#> qsec         1.28287213  1.28280223  1.28273392  1.28266717  1.28260197
#> vs          -0.82602500 -0.82604493 -0.82606349 -0.82608082 -0.82609703
#> wt          -6.19047983 -6.19007771 -6.18968519 -6.18930213 -6.18892836
#>                  0.0445      0.0415      0.0387      0.0361      0.0337
#> (Intercept)  8.32636818  8.32332967  8.32102592  8.31935996  8.31824393
#> am           5.51907228  5.51914879  5.51917512  5.51915787  5.51910304
#> carb         1.39753665  1.39724242  1.39696499  1.39670275  1.39645424
#> cyl          0.43007127  0.43015169  0.43019467  0.43020621  0.43019169
#> disp         0.03509135  0.03509023  0.03508894  0.03508747  0.03508583
#> drat         1.64704261  1.64654223  1.64608405  1.64566547  1.64528391
#> gear        -1.72357918 -1.72294561 -1.72238253 -1.72188252 -1.72143876
#> hp          -0.06096663 -0.06096082 -0.06095496 -0.06094909 -0.06094324
#> qsec         1.39530519  1.39538266  1.39543834  1.39547477  1.39549427
#> vs          -0.78264764 -0.78246772 -0.78232183 -0.78220621 -0.78211739
#> wt          -6.10156570 -6.10137310 -6.10117233 -6.10096346 -6.10074674
#>                  0.0314      0.0293      0.0273      0.0255      0.0237
#> (Intercept)  8.31759851  8.31735229  8.31744128  8.31780830  8.31840249
#> am           5.51901607  5.51890189  5.51876496  5.51860927  5.51843840
#> carb         1.39621813  1.39599321  1.39577840  1.39557272  1.39537529
#> cyl          0.43015589  0.43010303  0.43003683  0.42996055  0.42987704
#> disp         0.03508404  0.03508210  0.03508002  0.03507781  0.03507548
#> drat         1.64493682  1.64462175  1.64433632  1.64407825  1.64384535
#> gear        -1.72104496 -1.72069537 -1.72038475 -1.72010832 -1.71986177
#> hp          -0.06093743 -0.06093169 -0.06092603 -0.06092047 -0.06091502
#> qsec         1.39549897  1.39549080  1.39547155  1.39544281  1.39540604
#> vs          -0.78205217 -0.78200759 -0.78198096 -0.78196983 -0.78197197
#> wt          -6.10052247 -6.10029107 -6.10005300 -6.09980878 -6.09955897
#>                  0.0221      0.0206      0.0193      0.0180      0.0167
#> (Intercept)  8.31917880  8.32009750  8.32112376  8.32222718  8.32338142
#> am           5.51825554  5.51806350  5.51786480  5.51766163  5.51745590
#> carb         1.39518532  1.39500210  1.39482501  1.39465350  1.39448706
#> cyl          0.42978874  0.42969777  0.42960594  0.42951477  0.42942551
#> disp         0.03507305  0.03507052  0.03506791  0.03506523  0.03506248
#> drat         1.64363554  1.64344685  1.64327743  1.64312553  1.64298950
#> gear        -1.71964118 -1.71944304 -1.71926419 -1.71910183 -1.71895345
#> hp          -0.06090970 -0.06090451 -0.06089945 -0.06089454 -0.06088978
#> qsec         1.39536254  1.39531350  1.39525996  1.39520286  1.39514305
#> vs          -0.78198535 -0.78200815 -0.78203874 -0.78207566 -0.78211761
#> wt          -6.09930416 -6.09904495 -6.09878194 -6.09851574 -6.09824696
#>                  0.0156      0.0146      0.0136      0.0127      0.0118
#> (Intercept)  8.32456379  8.32575492  8.32693844  8.32810063  8.32923017
#> am           5.51724928  5.51704319  5.51683888  5.51663737  5.51643953
#> carb         1.39432527  1.39416775  1.39401415  1.39386420  1.39371764
#> cyl          0.42933923  0.42925675  0.42917876  0.42910576  0.42903814
#> disp         0.03505969  0.03505684  0.03505397  0.03505107  0.03504815
#> drat         1.64286784  1.64275911  1.64266199  1.64257529  1.64249787
#> gear        -1.71881683 -1.71869001 -1.71857127 -1.71845914 -1.71835230
#> hp          -0.06088516 -0.06088070 -0.06087638 -0.06087222 -0.06086820
#> qsec         1.39508124  1.39501810  1.39495418  1.39488999  1.39482593
#> vs          -0.78216344 -0.78221212 -0.78226279 -0.78231465 -0.78236705
#> wt          -6.09797618 -6.09770397 -6.09743089 -6.09715746 -6.09688418
#>                  0.0110      0.0103      0.0096      0.0089      0.0083
#> (Intercept)  8.33031787  8.33135640  8.33234013  8.33326486  8.33412766
#> am           5.51624607  5.51605758  5.51587451  5.51569721  5.51552595
#> carb         1.39357424  1.39343383  1.39329623  1.39316132  1.39302896
#> cyl          0.42897614  0.42891993  0.42886957  0.42882506  0.42878633
#> disp         0.03504523  0.03504230  0.03503937  0.03503646  0.03503356
#> drat         1.64242871  1.64236687  1.64231149  1.64226180  1.64221708
#> gear        -1.71824963 -1.71815019 -1.71805317 -1.71795788 -1.71786377
#> hp          -0.06086433 -0.06086059 -0.06085700 -0.06085354 -0.06085022
#> qsec         1.39476238  1.39469964  1.39463797  1.39457759  1.39451866
#> vs          -0.78241941 -0.78247124 -0.78252215 -0.78257178 -0.78261987
#> wt          -6.09661153 -6.09633996 -6.09606987 -6.09580166 -6.09553568
#>                  0.0078      0.0073      0.0068      0.0063      0.0059
#> (Intercept)  8.33492671  8.33566114  8.33633089  8.33693657  8.33747938
#> am           5.51536089  5.51520215  5.51504976  5.51490370  5.51476393
#> carb         1.39289908  1.39277157  1.39264637  1.39252342  1.39240268
#> cyl          0.42875325  0.42872565  0.42870334  0.42868609  0.42867365
#> disp         0.03503069  0.03502784  0.03502502  0.03502223  0.03501948
#> drat         1.64217670  1.64214009  1.64210672  1.64207615  1.64204796
#> gear        -1.71777038 -1.71767733 -1.71758434 -1.71749119 -1.71739772
#> hp          -0.06084701 -0.06084394 -0.06084098 -0.06083813 -0.06083539
#> qsec         1.39446133  1.39440571  1.39435186  1.39429986  1.39424972
#> vs          -0.78266619 -0.78271058 -0.78275291 -0.78279309 -0.78283106
#> wt          -6.09527226 -6.09501170 -6.09475425 -6.09450017 -6.09424967
#>                  0.0055      0.0051
#> (Intercept)  8.33796099  8.33838344
#> am           5.51463035  5.51450283
#> carb         1.39228409  1.39216763
#> cyl          0.42866574  0.42866211
#> disp         0.03501676  0.03501409
#> drat         1.64202178  1.64199731
#> gear        -1.71730384 -1.71720947
#> hp          -0.06083276 -0.06083023
#> qsec         1.39420147  1.39415511
#> vs          -0.78286679 -0.78290029
#> wt          -6.09400293 -6.09376013
#> 
#> $iter
#>   [1]   0   3   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   4   3   3   3   3   3   3   3   3   3   3   3   3   3   3  36  14
#>  [37]  28  28  27  15   1   1   1   1   1   1   1   1  88   1   1  44  47 149
#>  [55]   1   1  33 278   1   1   1   1   1   1   1   1   1   1  56   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.114955180 4.770224320 4.448727166 4.148897844 3.869276015 3.608499762
#>   [7] 3.365298955 3.138489069 2.926965410 2.729697737 2.545725246 2.374151885
#>  [13] 2.214141994 2.064916234 1.925747791 1.795958835 1.674917220 1.562033405
#>  [19] 1.456757582 1.358576996 1.267013454 1.181620987 1.101983688 1.027713675
#>  [25] 0.958449213 0.893852944 0.833610247 0.777427707 0.725031681 0.676166972
#>  [31] 0.630595582 0.588095550 0.548459878 0.511495518 0.477022432 0.444872717
#>  [37] 0.414889784 0.386927602 0.360849976 0.336529896 0.313848907 0.292696541
#>  [43] 0.272969774 0.254572525 0.237415188 0.221414199 0.206491623 0.192574779
#>  [49] 0.179595883 0.167491722 0.156203341 0.145675758 0.135857700 0.126701345
#>  [55] 0.118162099 0.110198369 0.102771368 0.095844921 0.089385294 0.083361025
#>  [61] 0.077742771 0.072503168 0.067616697 0.063059558 0.058809555 0.054845988
#>  [67] 0.051149552 0.047702243 0.044487272 0.041488978 0.038692760 0.036084998
#>  [73] 0.033652990 0.031384891 0.029269654 0.027296977 0.025457252 0.023741519
#>  [79] 0.022141420 0.020649162 0.019257478 0.017959588 0.016749172 0.015620334
#>  [85] 0.014567576 0.013585770 0.012670135 0.011816210 0.011019837 0.010277137
#>  [91] 0.009584492 0.008938529 0.008336102 0.007774277 0.007250317 0.006761670
#>  [97] 0.006305956 0.005880955 0.005484599 0.005114955
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 34
#> 
#> $loss
#>   [1] 749.7667 644.2950 556.0523 482.5591 421.6749 371.5527 330.5999 297.4442
#>   [9] 270.9040 249.9628 233.7470 221.5064 212.5977 206.4698 202.6514 200.7396
#>  [17] 200.3486 200.3486 200.3486 198.7197 191.0330 184.6178 179.2907 174.8930
#>  [25] 171.2877 168.3568 165.9987 164.1259 162.6631 161.5458 160.7185 160.1338
#>  [33] 159.7508 159.5350 141.7696 141.3578 136.6998 133.8017 132.3669 132.0866
#>  [41] 132.0866 132.0866 132.0866 132.0866 132.0866 132.0866 132.0866 132.0866
#>  [49] 119.2355 119.2355 119.2355 118.4940 118.0126 114.3005 114.3004 114.3004
#>  [57] 114.0333 108.9589 108.9589 108.9588 108.9588 108.9588 108.9588 108.9588
#>  [65] 108.9588 108.9587 108.9587 108.9587 108.2409 108.2409 108.2409 108.2409
#>  [73] 108.2409 108.2408 108.2408 108.2408 108.2408 108.2408 108.2408 108.2408
#>  [81] 108.2408 108.2408 108.2407 108.2407 108.2407 108.2407 108.2407 108.2407
#>  [89] 108.2407 108.2407 108.2407 108.2407 108.2407 108.2407 108.2407 108.2406
#>  [97] 108.2406 108.2406 108.2406 108.2406
#> 
#> $linear.predictors
#>      5.1150   4.7702   4.4487   4.1489   3.8693   3.6085   3.3653   3.1385
#> 1  21.03333 21.44783 21.83440 22.19491 22.53112 22.84467 23.13709 23.40981
#> 2  21.03333 20.98650 20.94282 20.90209 20.86410 20.82868 20.79564 20.76482
#> 3  21.03333 20.86021 20.69876 20.54819 20.40777 20.27681 20.15468 20.04078
#> 4  21.03333 21.02001 21.00758 20.99598 20.98517 20.97509 20.96569 20.95692
#> 5  21.03333 20.87052 20.71869 20.57708 20.44502 20.32186 20.20700 20.09988
#> 6  21.03333 20.87052 20.71869 20.57708 20.44502 20.32186 20.20700 20.09988
#> 7  21.03333 20.54579 20.09110 19.66706 19.27159 18.90278 18.55883 18.23805
#> 8  21.03333 19.84786 18.74228 17.71122 16.74965 15.85288 15.01655 14.23659
#> 9  21.03333 19.88858 18.82098 17.82534 16.89679 16.03083 15.22323 14.47006
#> 10 21.03333 21.50969 21.95394 22.36825 22.75463 23.11497 23.45103 23.76444
#> 11 21.03333 21.69783 22.31754 22.89548 23.43447 23.93714 24.40593 24.84312
#> 12 21.03333 21.37309 21.68995 21.98546 22.26104 22.51806 22.75775 22.98129
#> 13 21.03333 20.82929 20.63899 20.46152 20.29601 20.14166 19.99771 19.86346
#> 14 21.03333 20.66176 20.31524 19.99207 19.69067 19.40960 19.14746 18.90299
#> 15 21.03333 21.64628 22.21792 22.75103 23.24822 23.71189 24.14431 24.54759
#> 16 21.03333 21.54061 22.01371 22.45491 22.86639 23.25013 23.60800 23.94176
#> 17 21.03333 21.86380 22.63831 23.36061 24.03423 24.66245 25.24833 25.79473
#> 18 21.03333 21.00970 20.98765 20.96709 20.94792 20.93004 20.91336 20.89781
#> 19 21.03333 21.21588 21.38612 21.54489 21.69296 21.83104 21.95983 22.07993
#> 20 21.03333 20.80351 20.58918 20.38930 20.20289 20.02904 19.86690 19.71570
#> 21 21.03333 21.21072 21.37616 21.53044 21.67433 21.80852 21.93366 22.05037
#>      2.9270   2.7297   2.5457   2.3742   2.2141   2.0649   1.9257    1.7960
#> 1  23.66414 23.90133 24.12254 24.32884 24.52123 24.70066 24.86799 25.024046
#> 2  20.73609 20.70929 20.68430 20.66099 20.63925 20.61898 20.60007 20.582437
#> 3  19.93455 19.83549 19.74310 19.65694 19.57658 19.50164 19.43176 19.366578
#> 4  20.94874 20.94111 20.93400 20.92737 20.92118 20.91541 20.91003 20.905012
#> 5  19.99999 19.90682 19.81993 19.73890 19.66333 19.59286 19.52713 19.465832
#> 6  19.99999 19.90682 19.81993 19.73890 19.66333 19.59286 19.52713 19.465832
#> 7  17.93890 17.65991 17.39972 17.15706 16.93076 16.71972 16.52289 16.339336
#> 8  13.50920 12.83083 12.19817 11.60816 11.05791 10.54475 10.06617  9.619851
#> 9  13.76765 13.11258 12.50166 11.93192 11.40057 10.90503 10.44290 10.011904
#> 10 24.05673 24.32932 24.58353 24.82061 25.04172 25.24792 25.44023 25.619569
#> 11 25.25085 25.63110 25.98572 26.31644 26.62487 26.91252 27.18077 27.430952
#> 12 23.18976 23.38419 23.56550 23.73460 23.89230 24.03938 24.17654 24.304456
#> 13 19.73826 19.62150 19.51260 19.41105 19.31634 19.22801 19.14564 19.068816
#> 14 18.67500 18.46238 18.26408 18.07915 17.90668 17.74584 17.59584 17.455942
#> 15 24.92369 25.27445 25.60156 25.90663 26.19113 26.45646 26.70391 26.934683
#> 16 24.25302 24.54331 24.81403 25.06650 25.30196 25.52155 25.72634 25.917331
#> 17 26.30430 26.77952 27.22272 27.63605 28.02152 28.38101 28.71627 29.028939
#> 18 20.88331 20.86978 20.85717 20.84540 20.83443 20.82420 20.81466 20.805758
#> 19 22.19194 22.29639 22.39381 22.48467 22.56940 22.64841 22.72211 22.790835
#> 20 19.57468 19.44317 19.32052 19.20614 19.09947 18.99999 18.90721 18.820682
#> 21 22.15922 22.26073 22.35540 22.44368 22.52602 22.60281 22.67442 22.741208
#>       1.6749    1.5620    1.4568    1.3586    1.2670    1.1816    1.1020
#> 1  25.133432 25.133432 25.133432 25.137261 25.156511 25.174464 25.191207
#> 2  20.570078 20.570078 20.570078 20.590228 20.691517 20.785984 20.874085
#> 3  19.320892 19.320892 19.320892 19.353622 19.518152 19.671602 19.814710
#> 4  20.901495 20.901495 20.901495 20.972124 21.327170 21.658305 21.967122
#> 5  19.422866 19.422866 19.422866 19.427325 19.449739 19.470644 19.490139
#> 6  19.422866 19.422866 19.422866 19.436131 19.502811 19.565000 19.622998
#> 7  16.210673 16.210673 16.210673 16.204837 16.175496 16.148132 16.122612
#> 8   9.307008  9.307008  9.307008  9.313596  9.346717  9.377607  9.406415
#> 9   9.709806  9.709806  9.709806  9.710159  9.711933  9.713587  9.715130
#> 10 25.745278 25.745278 25.745278 25.761174 25.841082 25.915609 25.985112
#> 11 27.606311 27.606311 27.606311 27.626830 27.729977 27.826178 27.915895
#> 12 24.394118 24.394118 24.394118 24.419165 24.545072 24.662500 24.772014
#> 13 19.014969 19.014969 19.014969 18.998810 18.917583 18.841826 18.771175
#> 14 17.357885 17.357885 17.357885 17.345871 17.285479 17.229153 17.176624
#> 15 27.096439 27.096439 27.096439 27.102744 27.134439 27.163999 27.191568
#> 16 26.051201 26.051201 26.051201 26.026166 25.900318 25.782945 25.673482
#> 17 29.248098 29.248098 29.248098 29.223099 29.097431 28.980226 28.870920
#> 18 20.799521 20.799521 20.799521 20.746960 20.482746 20.236326 20.006513
#> 19 22.839008 22.839008 22.839008 22.799275 22.599539 22.413255 22.239525
#> 20 18.760033 18.760033 18.760033 18.710790 18.463252 18.232385 18.017077
#> 21 22.788021 22.788021 22.788021 22.793831 22.823035 22.850273 22.875676
#>       1.0277    0.9584    0.8939    0.8336    0.7774    0.7250    0.6762
#> 1  25.206821 25.221384 25.234964 25.247630 25.259442 25.270458 25.280731
#> 2  20.956249 21.032874 21.104336 21.170981 21.233134 21.291099 21.345156
#> 3  19.948172 20.072640 20.188719 20.296975 20.397935 20.492090 20.579900
#> 4  22.255126 22.523720 22.774211 23.007821 23.225685 23.428867 23.618354
#> 5  19.508321 19.525277 19.541090 19.555838 19.569592 19.582418 19.594381
#> 6  19.677087 19.727531 19.774575 19.818448 19.859365 19.897524 19.933111
#> 7  16.098812 16.076616 16.055915 16.036610 16.018606 16.001816 15.986157
#> 8   9.433282  9.458338  9.481705  9.503498  9.523821  9.542775  9.560452
#> 9   9.716568  9.717910  9.719161  9.720328  9.721417  9.722431  9.723378
#> 10 26.049932 26.110382 26.166759 26.219336 26.268369 26.314098 26.356745
#> 11 27.999565 28.077597 28.150369 28.218236 28.281530 28.340558 28.395608
#> 12 24.874147 24.969397 25.058227 25.141070 25.218330 25.290383 25.357579
#> 13 18.705286 18.643837 18.586530 18.533085 18.483242 18.436758 18.393407
#> 14 17.127635 17.081948 17.039340 16.999604 16.962545 16.927985 16.895753
#> 15 27.217278 27.241255 27.263617 27.284471 27.303920 27.322058 27.338974
#> 16 25.571397 25.476192 25.387404 25.304599 25.227376 25.155357 25.088192
#> 17 28.768981 28.673913 28.585251 28.502566 28.425453 28.353537 28.286468
#> 18 19.792190 19.592310 19.405902 19.232058 19.069929 18.918728 18.777717
#> 19 22.077505 21.926404 21.785487 21.654067 21.531504 21.417202 21.310603
#> 20 17.816280 17.629016 17.454373 17.291501 17.139606 16.997947 16.865837
#> 21 22.899366 22.921459 22.942064 22.961279 22.979200 22.995913 23.011499
#>       0.6306    0.5881    0.5485    0.5115   0.4770   0.4449   0.4149   0.3869
#> 1  25.290312 25.299247 25.307580 25.315352 26.24470 26.25870 26.44248 26.61431
#> 2  21.395571 21.442588 21.486436 21.527329 20.69124 20.69950 20.81402 20.92095
#> 3  20.661791 20.738164 20.809389 20.875814 20.66272 20.66975 20.76513 20.85424
#> 4  23.795071 23.959877 24.113577 24.256917 24.81439 24.79123 24.47866 24.18665
#> 5  19.605537 19.615941 19.625644 19.634693 18.58160 18.56246 18.31035 18.07466
#> 6  19.966299 19.997251 20.026117 20.053037 19.25501 19.23265 18.93633 18.65935
#> 7  15.971553 15.957934 15.945233 15.933387 15.24703 15.25239 15.32714 15.39692
#> 8   9.576937  9.592311  9.606649  9.620021 10.72270 10.73004 10.82378 10.91149
#> 9   9.724261  9.725084  9.725852  9.726568 10.56524 10.57377 10.68463 10.78830
#> 10 26.396517 26.433609 26.468201 26.500462 27.65267 27.66061 27.76224 27.85733
#> 11 28.446947 28.494826 28.539479 28.581122 29.48199 29.48322 29.49605 29.50812
#> 12 25.420247 25.478691 25.533197 25.584029 24.09818 24.09435 24.04949 24.00744
#> 13 18.352978 18.315273 18.280110 18.247317 16.68147 16.69512 16.88481 17.06190
#> 14 16.865694 16.837661 16.811517 16.787135 15.68437 15.70097 15.92826 16.14052
#> 15 27.354749 27.369461 27.383182 27.395978 27.99068 27.99848 28.10065 28.19617
#> 16 25.025553 24.967136 24.912656 24.861848 24.76514 24.77514 24.90956 25.03515
#> 17 28.223919 28.165586 28.111184 28.060449 27.30299 27.30437 27.32592 27.34599
#> 18 18.646210 18.523566 18.409188 18.302518 18.49568 18.50575 18.63957 18.76463
#> 19 21.211188 21.118474 21.032009 20.951371 21.09387 21.06960 20.74324 20.43828
#> 20 16.742629 16.627726 16.520567 16.420630 17.13208 17.09812 16.63937 16.21076
#> 21 23.026036 23.039592 23.052235 23.064025 24.53626 24.54378 24.63833 24.72684
#>      0.3608   0.3365   0.3138   0.2927   0.2730   0.2546   0.2374   0.2214
#> 1  26.77418 26.88225 26.88252 26.88272 26.88286 26.88296 26.88303 26.88307
#> 2  21.02054 21.08822 21.08807 21.08796 21.08788 21.08782 21.08778 21.08775
#> 3  20.93720 20.99346 20.99345 20.99344 20.99344 20.99343 20.99343 20.99342
#> 4  23.91480 23.73053 23.73058 23.73062 23.73065 23.73067 23.73069 23.73070
#> 5  17.85537 17.70708 17.70675 17.70652 17.70636 17.70624 17.70616 17.70610
#> 6  18.40161 18.22720 18.22693 18.22673 18.22659 18.22650 18.22643 18.22638
#> 7  15.46193 15.50613 15.50599 15.50589 15.50582 15.50577 15.50573 15.50570
#> 8  10.99304 11.04802 11.04831 11.04851 11.04866 11.04876 11.04884 11.04889
#> 9  10.88473 10.94985 10.95008 10.95025 10.95036 10.95044 10.95050 10.95055
#> 10 27.94574 28.00536 28.00567 28.00589 28.00604 28.00615 28.00623 28.00628
#> 11 29.51929 29.52665 29.52687 29.52703 29.52714 29.52722 29.52728 29.52732
#> 12 23.96840 23.94233 23.94198 23.94173 23.94154 23.94142 23.94132 23.94126
#> 13 17.22685 17.33897 17.33866 17.33844 17.33828 17.33816 17.33808 17.33802
#> 14 16.33818 16.47235 16.47216 16.47203 16.47193 16.47185 16.47180 16.47176
#> 15 28.28504 28.34510 28.34528 28.34540 28.34549 28.34555 28.34559 28.34562
#> 16 25.15205 25.23126 25.23128 25.23129 25.23129 25.23129 25.23129 25.23130
#> 17 27.36472 27.37759 27.37741 27.37729 27.37720 27.37713 27.37708 27.37705
#> 18 18.88102 18.95980 18.95988 18.95993 18.95996 18.95999 18.96001 18.96002
#> 19 20.15442 19.96207 19.96199 19.96193 19.96190 19.96187 19.96186 19.96185
#> 20 15.81177 15.54129 15.54129 15.54130 15.54131 15.54132 15.54133 15.54134
#> 21 24.80911 24.86448 24.86486 24.86513 24.86532 24.86545 24.86555 24.86562
#>      0.2065   0.1926   0.1796   0.1675   0.1562   0.1457   0.1359   0.1267
#> 1  26.88311 26.88313 26.68592 26.68577 26.68564 26.62352 26.54499 26.02116
#> 2  21.08773 21.08771 21.38680 21.38692 21.38702 21.32936 21.25872 21.59257
#> 3  20.99342 20.99342 20.81406 20.81409 20.81411 20.63464 20.40842 20.49025
#> 4  23.73071 23.73072 23.24679 23.24712 23.24740 23.31967 23.41058 23.79197
#> 5  17.70606 17.70603 17.85178 17.85162 17.85149 17.97072 18.11847 18.19667
#> 6  18.22635 18.22632 18.35789 18.35780 18.35774 18.47131 18.61220 18.88053
#> 7  15.50568 15.50567 14.66907 14.66900 14.66895 14.60174 14.51492 13.53262
#> 8  11.04893 11.04895 11.78712 11.78692 11.78676 11.77766 11.76717 12.21889
#> 9  10.95057 10.95060 10.97451 10.97437 10.97425 11.01741 11.07194 10.68307
#> 10 28.00632 28.00635 28.24028 28.24008 28.23991 28.22450 28.20505 28.13661
#> 11 29.52735 29.52737 29.84588 29.84589 29.84589 29.86854 29.89823 30.23390
#> 12 23.94121 23.94118 22.92557 22.92585 22.92609 22.96513 23.01391 22.21649
#> 13 17.33798 17.33795 17.77750 17.77748 17.77747 17.67670 17.55116 17.76622
#> 14 16.47174 16.47172 17.20740 17.20751 17.20761 17.23383 17.27077 17.63130
#> 15 28.34564 28.34566 28.72669 28.72654 28.72642 28.72229 28.71767 28.80086
#> 16 25.23130 25.23130 25.97478 25.97447 25.97422 26.10388 26.26788 26.11312
#> 17 27.37703 27.37701 27.36538 27.36550 27.36560 27.30083 27.21979 27.40574
#> 18 18.96003 18.96003 18.50328 18.50362 18.50390 18.67689 18.89790 18.18243
#> 19 19.96184 19.96183 20.45491 20.45473 20.45458 20.34979 20.21457 21.32394
#> 20 15.54134 15.54135 14.19635 14.19690 14.19736 14.11953 14.01914 14.32143
#> 21 24.86567 24.86570 24.70807 24.70781 24.70760 24.71207 24.71651 24.16023
#>      0.1182   0.1102   0.1028   0.0958   0.0894   0.0834   0.0777   0.0725
#> 1  26.02103 26.02090 25.99936 25.61814 25.61827 25.61839 25.61850 25.61862
#> 2  21.59268 21.59279 21.56259 21.32751 21.32740 21.32728 21.32718 21.32707
#> 3  20.49020 20.49016 20.40669 19.36547 19.36556 19.36565 19.36575 19.36584
#> 4  23.79212 23.79226 23.82190 23.48470 23.48470 23.48471 23.48471 23.48471
#> 5  18.19661 18.19654 18.24756 18.10034 18.10039 18.10043 18.10048 18.10052
#> 6  18.88052 18.88051 18.92709 18.87034 18.87034 18.87033 18.87033 18.87033
#> 7  13.53233 13.53205 13.51202 13.68889 13.68914 13.68939 13.68963 13.68986
#> 8  12.21899 12.21908 12.20994 12.13314 12.13305 12.13296 12.13288 12.13280
#> 9  10.68301 10.68295 10.70808 10.80442 10.80447 10.80452 10.80457 10.80462
#> 10 28.13658 28.13655 28.13074 28.31551 28.31550 28.31550 28.31549 28.31549
#> 11 30.23403 30.23414 30.24157 31.11274 31.11256 31.11237 31.11219 31.11202
#> 12 22.21636 22.21624 22.24364 23.18519 23.18523 23.18526 23.18529 23.18531
#> 13 17.76625 17.76627 17.71655 17.85316 17.85311 17.85307 17.85303 17.85299
#> 14 17.63150 17.63169 17.64023 18.18570 18.18549 18.18529 18.18510 18.18491
#> 15 28.80088 28.80090 28.79827 29.23659 29.23651 29.23643 29.23635 29.23627
#> 16 26.11310 26.11308 26.17371 26.17193 26.17205 26.17216 26.17227 26.17238
#> 17 27.40578 27.40582 27.37378 26.20371 26.20376 26.20381 26.20387 26.20392
#> 18 18.18250 18.18257 18.27366 18.05539 18.05542 18.05545 18.05547 18.05550
#> 19 21.32399 21.32404 21.26005 21.34948 21.34941 21.34933 21.34925 21.34918
#> 20 14.32149 14.32155 14.28350 14.46060 14.46051 14.46042 14.46033 14.46024
#> 21 24.16007 24.15992 24.16907 24.17704 24.17714 24.17723 24.17733 24.17742
#>      0.0676   0.0631   0.0588   0.0548   0.0511   0.0477   0.0445   0.0415
#> 1  25.61873 25.61884 25.61895 25.61905 25.61915 25.61925 25.55838 25.55854
#> 2  21.32697 21.32687 21.32677 21.32668 21.32658 21.32650 21.27928 21.27944
#> 3  19.36592 19.36601 19.36610 19.36618 19.36626 19.36635 19.49250 19.49294
#> 4  23.48471 23.48471 23.48472 23.48472 23.48472 23.48472 23.50276 23.50298
#> 5  18.10057 18.10061 18.10066 18.10070 18.10074 18.10078 18.20350 18.20313
#> 6  18.87034 18.87034 18.87034 18.87034 18.87034 18.87034 19.04068 19.04036
#> 7  13.69009 13.69031 13.69052 13.69073 13.69093 13.69112 13.99437 13.99433
#> 8  12.13272 12.13264 12.13257 12.13250 12.13243 12.13236 11.93112 11.93110
#> 9  10.80466 10.80471 10.80475 10.80479 10.80484 10.80488 10.61751 10.61745
#> 10 28.31549 28.31548 28.31548 28.31547 28.31547 28.31546 28.48727 28.48724
#> 11 31.11184 31.11167 31.11151 31.11135 31.11119 31.11103 31.33918 31.33904
#> 12 23.18534 23.18536 23.18538 23.18540 23.18542 23.18544 22.76527 22.76494
#> 13 17.85295 17.85292 17.85288 17.85285 17.85281 17.85278 18.01245 18.01249
#> 14 18.18472 18.18454 18.18436 18.18419 18.18402 18.18385 18.16098 18.16099
#> 15 29.23619 29.23612 29.23605 29.23598 29.23591 29.23584 29.31939 29.31926
#> 16 26.17249 26.17259 26.17269 26.17279 26.17288 26.17298 25.95707 25.95690
#> 17 26.20397 26.20402 26.20408 26.20413 26.20418 26.20423 25.96655 25.96694
#> 18 18.05553 18.05555 18.05558 18.05560 18.05562 18.05565 18.32062 18.32105
#> 19 21.34911 21.34904 21.34898 21.34891 21.34885 21.34879 21.30046 21.30017
#> 20 14.46015 14.46007 14.45998 14.45990 14.45982 14.45974 14.40648 14.40663
#> 21 24.17751 24.17759 24.17768 24.17776 24.17784 24.17792 24.04419 24.04410
#>      0.0387   0.0361   0.0337   0.0314   0.0293   0.0273   0.0255   0.0237
#> 1  25.55870 25.55885 25.55900 25.55915 25.55929 25.55943 25.55956 25.55969
#> 2  21.27958 21.27969 21.27979 21.27986 21.27991 21.27994 21.27996 21.27996
#> 3  19.49333 19.49367 19.49398 19.49425 19.49449 19.49470 19.49488 19.49504
#> 4  23.50318 23.50335 23.50350 23.50364 23.50375 23.50385 23.50393 23.50400
#> 5  18.20279 18.20248 18.20219 18.20194 18.20171 18.20151 18.20133 18.20117
#> 6  19.04005 19.03976 19.03949 19.03924 19.03901 19.03879 19.03860 19.03841
#> 7  13.99430 13.99429 13.99430 13.99432 13.99435 13.99440 13.99446 13.99454
#> 8  11.93109 11.93108 11.93107 11.93106 11.93105 11.93103 11.93101 11.93099
#> 9  10.61741 10.61739 10.61739 10.61740 10.61742 10.61745 10.61749 10.61754
#> 10 28.48719 28.48714 28.48709 28.48703 28.48696 28.48690 28.48683 28.48677
#> 11 31.33889 31.33873 31.33855 31.33837 31.33818 31.33800 31.33780 31.33761
#> 12 22.76469 22.76451 22.76440 22.76433 22.76430 22.76432 22.76436 22.76443
#> 13 18.01251 18.01252 18.01252 18.01251 18.01249 18.01247 18.01243 18.01240
#> 14 18.16098 18.16095 18.16092 18.16087 18.16080 18.16073 18.16065 18.16056
#> 15 29.31913 29.31900 29.31888 29.31875 29.31863 29.31852 29.31840 29.31829
#> 16 25.95676 25.95666 25.95658 25.95654 25.95652 25.95652 25.95653 25.95656
#> 17 25.96729 25.96762 25.96793 25.96821 25.96847 25.96871 25.96893 25.96913
#> 18 18.32143 18.32175 18.32201 18.32224 18.32242 18.32257 18.32269 18.32279
#> 19 21.29990 21.29965 21.29943 21.29923 21.29904 21.29888 21.29873 21.29860
#> 20 14.40678 14.40691 14.40703 14.40713 14.40723 14.40731 14.40739 14.40745
#> 21 24.04403 24.04398 24.04396 24.04395 24.04396 24.04398 24.04401 24.04406
#>      0.0221   0.0206   0.0193   0.0180   0.0167   0.0156   0.0146   0.0136
#> 1  25.55982 25.55994 25.56006 25.56017 25.56028 25.56039 25.56049 25.56059
#> 2  21.27995 21.27993 21.27990 21.27986 21.27981 21.27976 21.27970 21.27964
#> 3  19.49518 19.49530 19.49541 19.49550 19.49558 19.49566 19.49572 19.49578
#> 4  23.50406 23.50411 23.50415 23.50419 23.50421 23.50423 23.50425 23.50425
#> 5  18.20103 18.20091 18.20081 18.20072 18.20065 18.20059 18.20054 18.20050
#> 6  19.03825 19.03810 19.03796 19.03784 19.03773 19.03764 19.03755 19.03748
#> 7  13.99462 13.99472 13.99483 13.99494 13.99506 13.99519 13.99532 13.99546
#> 8  11.93097 11.93094 11.93091 11.93088 11.93084 11.93080 11.93076 11.93072
#> 9  10.61760 10.61765 10.61771 10.61777 10.61784 10.61790 10.61796 10.61802
#> 10 28.48671 28.48664 28.48659 28.48653 28.48648 28.48643 28.48638 28.48634
#> 11 31.33742 31.33723 31.33705 31.33686 31.33668 31.33651 31.33633 31.33617
#> 12 22.76451 22.76462 22.76473 22.76485 22.76498 22.76511 22.76525 22.76538
#> 13 18.01236 18.01232 18.01228 18.01223 18.01219 18.01214 18.01210 18.01206
#> 14 18.16046 18.16036 18.16025 18.16014 18.16002 18.15990 18.15977 18.15965
#> 15 29.31819 29.31808 29.31799 29.31789 29.31780 29.31771 29.31763 29.31755
#> 16 25.95661 25.95666 25.95673 25.95680 25.95688 25.95696 25.95704 25.95713
#> 17 25.96931 25.96947 25.96962 25.96975 25.96987 25.96997 25.97007 25.97015
#> 18 18.32286 18.32292 18.32296 18.32298 18.32300 18.32301 18.32301 18.32300
#> 19 21.29848 21.29836 21.29826 21.29817 21.29809 21.29801 21.29794 21.29788
#> 20 14.40751 14.40755 14.40759 14.40761 14.40763 14.40764 14.40765 14.40765
#> 21 24.04411 24.04417 24.04424 24.04431 24.04438 24.04445 24.04453 24.04461
#>      0.0127   0.0118   0.0110   0.0103   0.0096   0.0089   0.0083   0.0078
#> 1  25.56069 25.56078 25.56087 25.56095 25.56104 25.56112 25.56119 25.56127
#> 2  21.27957 21.27950 21.27943 21.27935 21.27928 21.27920 21.27912 21.27905
#> 3  19.49583 19.49588 19.49593 19.49597 19.49600 19.49604 19.49608 19.49611
#> 4  23.50426 23.50426 23.50426 23.50426 23.50426 23.50425 23.50424 23.50423
#> 5  18.20048 18.20046 18.20045 18.20045 18.20045 18.20046 18.20048 18.20049
#> 6  19.03741 19.03736 19.03731 19.03727 19.03723 19.03721 19.03719 19.03717
#> 7  13.99561 13.99576 13.99591 13.99606 13.99622 13.99637 13.99653 13.99669
#> 8  11.93067 11.93063 11.93058 11.93053 11.93048 11.93043 11.93037 11.93032
#> 9  10.61808 10.61814 10.61820 10.61825 10.61830 10.61835 10.61840 10.61844
#> 10 28.48630 28.48626 28.48623 28.48620 28.48617 28.48615 28.48613 28.48611
#> 11 31.33601 31.33585 31.33570 31.33555 31.33541 31.33528 31.33514 31.33502
#> 12 22.76551 22.76563 22.76575 22.76587 22.76598 22.76608 22.76618 22.76627
#> 13 18.01201 18.01197 18.01193 18.01189 18.01186 18.01182 18.01179 18.01176
#> 14 18.15952 18.15940 18.15927 18.15914 18.15901 18.15889 18.15876 18.15864
#> 15 29.31747 29.31740 29.31733 29.31726 29.31720 29.31714 29.31708 29.31702
#> 16 25.95722 25.95731 25.95740 25.95749 25.95757 25.95766 25.95774 25.95782
#> 17 25.97023 25.97029 25.97035 25.97040 25.97044 25.97048 25.97051 25.97054
#> 18 18.32299 18.32297 18.32296 18.32294 18.32293 18.32291 18.32289 18.32288
#> 19 21.29782 21.29777 21.29771 21.29767 21.29762 21.29758 21.29754 21.29750
#> 20 14.40764 14.40763 14.40761 14.40759 14.40756 14.40753 14.40750 14.40746
#> 21 24.04469 24.04476 24.04484 24.04492 24.04499 24.04507 24.04514 24.04521
#>      0.0073   0.0068   0.0063   0.0059   0.0055   0.0051
#> 1  25.56134 25.56141 25.56148 25.56154 25.56161 25.56167
#> 2  21.27897 21.27889 21.27882 21.27874 21.27867 21.27859
#> 3  19.49614 19.49618 19.49621 19.49624 19.49628 19.49631
#> 4  23.50423 23.50422 23.50421 23.50420 23.50419 23.50418
#> 5  18.20052 18.20054 18.20057 18.20060 18.20063 18.20067
#> 6  19.03716 19.03715 19.03715 19.03715 19.03715 19.03716
#> 7  13.99685 13.99701 13.99716 13.99732 13.99748 13.99763
#> 8  11.93027 11.93021 11.93016 11.93010 11.93004 11.92999
#> 9  10.61849 10.61853 10.61856 10.61860 10.61863 10.61866
#> 10 28.48609 28.48608 28.48607 28.48606 28.48606 28.48606
#> 11 31.33490 31.33478 31.33467 31.33457 31.33446 31.33437
#> 12 22.76635 22.76642 22.76649 22.76655 22.76660 22.76665
#> 13 18.01173 18.01170 18.01168 18.01165 18.01163 18.01161
#> 14 18.15851 18.15839 18.15827 18.15815 18.15804 18.15792
#> 15 29.31697 29.31692 29.31687 29.31682 29.31678 29.31674
#> 16 25.95790 25.95798 25.95805 25.95813 25.95819 25.95826
#> 17 25.97057 25.97059 25.97060 25.97062 25.97063 25.97064
#> 18 18.32287 18.32286 18.32285 18.32284 18.32283 18.32283
#> 19 21.29746 21.29742 21.29739 21.29736 21.29732 21.29729
#> 20 14.40742 14.40738 14.40734 14.40730 14.40725 14.40720
#> 21 24.04528 24.04534 24.04541 24.04547 24.04553 24.04558
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 22.8 21.4 18.1 22.8 19.2 17.8 16.4 10.4 14.7 32.4 33.9 21.5 15.5 19.2 27.3
#> [16] 26.0 30.4 15.8 19.7 15.0 21.4
#> 
#> $X
#>               am       carb        cyl       disp        drat       gear
#>  [1,]  1.0488088 -0.9085547 -1.0427605 -0.8496288  0.42585861  0.1846372
#>  [2,] -0.9534626 -0.9085547  0.1097643  0.3808115 -1.10780146 -1.1078234
#>  [3,] -0.9534626 -0.9085547  0.1097643  0.1101146 -1.74516668 -1.1078234
#>  [4,] -0.9534626 -0.3634219 -1.0427605 -0.5805725  0.56528225  0.1846372
#>  [5,] -0.9534626  0.7268438  0.1097643 -0.3607338  0.56528225  0.1846372
#>  [6,] -0.9534626  0.7268438  0.1097643 -0.3607338  0.56528225  0.1846372
#>  [7,] -0.9534626  0.1817109  1.2622890  0.5268237 -1.12771912 -1.1078234
#>  [8,] -0.9534626  0.7268438  1.2622890  2.0378044 -1.26714276 -1.1078234
#>  [9,] -0.9534626  0.7268438  1.2622890  1.8737457 -0.80903651 -1.1078234
#> [10,]  1.0488088 -0.9085547 -1.0427605 -1.0899748  0.88396486  0.1846372
#> [11,]  1.0488088 -0.9085547 -1.0427605 -1.1523171  1.16281215  0.1846372
#> [12,] -0.9534626 -0.9085547 -1.0427605 -0.7503733  0.12709366 -1.1078234
#> [13,] -0.9534626 -0.3634219  1.2622890  0.8729876 -1.74516668 -1.1078234
#> [14,] -0.9534626 -0.3634219  1.2622890  1.5456283 -1.10780146 -1.1078234
#> [15,]  1.0488088 -0.9085547 -1.0427605 -1.0875139  0.88396486  0.1846372
#> [16,]  1.0488088 -0.3634219 -1.0427605 -0.7487327  1.58108307  1.4770979
#> [17,]  1.0488088 -0.3634219 -1.0427605 -0.9554466  0.26651730  1.4770979
#> [18,]  1.0488088  0.7268438  1.2622890  1.1436845  1.16281215  1.4770979
#> [19,]  1.0488088  1.8171095  0.1097643 -0.5461202 -0.03224765  1.4770979
#> [20,]  1.0488088  2.9073751  1.2622890  0.7335377 -0.19158895  1.4770979
#> [21,]  1.0488088 -0.3634219 -1.0427605 -0.7429906  0.94371785  0.1846372
#>               hp       qsec         vs          wt
#>  [1,] -0.7063919  0.2652547  0.8660254 -0.80159033
#>  [2,] -0.4608399  0.6933462  0.8660254  0.09056874
#>  [3,] -0.5330610  1.0956491  0.8660254  0.33479106
#>  [4,] -0.6775034  2.4779207  0.8660254  0.02577507
#>  [5,] -0.2730648  0.1053650  0.8660254  0.31485454
#>  [6,] -0.2730648  0.4148288  0.8660254  0.31485454
#>  [7,]  0.5502566 -0.3588306 -1.1547005  0.94285479
#>  [8,]  1.0558048 -0.1422060 -1.1547005  2.29255690
#>  [9,]  1.2724683 -0.3485152 -1.1547005  2.21380766
#> [10,] -1.0963862  0.7088194  0.8660254 -0.92120943
#> [11,] -1.1108304  0.9306018  0.8660254 -1.28505084
#> [12,] -0.6486149  0.9873368  0.8660254 -0.65705059
#> [13,]  0.1169295 -0.6321903 -1.1547005  0.39460061
#> [14,]  0.4780354 -0.5393512 -1.1547005  0.71856899
#> [15,] -1.0963862  0.4148288  0.8660254 -1.18536826
#> [16,] -0.7352803 -0.7198717 -1.1547005 -0.98101897
#> [17,] -0.4175072 -0.6167171  0.8660254 -1.60602874
#> [18,]  1.7635723 -1.8545722 -1.1547005  0.04571158
#> [19,]  0.4780354 -1.3387992 -1.1547005 -0.35301873
#> [20,]  2.7891130 -1.8029949 -1.1547005  0.44444190
#> [21,] -0.4752841  0.2600969  0.8660254 -0.34305047
#> attr(,"center")
#>  [1]   0.4761905   2.6666667   5.8095238 211.5761905   3.6361905   3.8571429
#>  [7] 141.9047619  18.0957143   0.5714286   3.1241429
#> attr(,"scale")
#>  [1]   0.4994328   1.8344153   1.7353207 121.9075837   0.5020669   0.7737179
#>  [7]  69.2317741   1.9388376   0.4948717   1.0031843
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 14
#> 
#> $lambda.min
#> [1] 2.064916
#> 
#> $null.dev
#> [1] 35.70317
#> 
#> $Bias
#> [1] 8.264316
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#>  8.97097