Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 48.475293 45.421968 39.330274 33.818511 29.283920 25.581750 22.587283
#>   [8] 20.193142 18.306960 16.849358 15.752185 14.957000 14.413746 14.079611
#>  [15] 13.918031 13.896907 13.690341 13.597892 13.595881 13.565683 13.524976
#>  [22] 13.499668 13.487067 13.484761 13.474684 13.337402 13.187146 14.156394
#>  [29] 14.252827 14.206822 14.181459 13.209480 12.155173 11.304625  9.964999
#>  [36]  9.995662  9.992567  9.992500  9.992454  9.992421 18.521266 23.956500
#>  [43] 23.477427 24.098990 24.195996 24.981532 25.000029 25.660119 25.830270
#>  [50] 26.126827 25.893648 25.674790 25.562025 26.148917 27.666766 30.360521
#>  [57] 30.798709 30.799952 28.408980 28.408438 39.170562 39.269579 39.269893
#>  [64] 39.324075 39.811111 40.432431 41.433292 40.953714 40.955223 41.446063
#>  [71] 41.447498 41.448883 41.450233 39.114894 39.098853 39.097580 39.096380
#>  [78] 39.066579 39.065558 39.056641 39.055767 38.984986 38.969000 38.954703
#>  [85] 38.953941 38.953255 38.952640 38.952093 38.951934 39.017384 38.681595
#>  [92] 38.676348 38.654793 38.649825 38.645002 38.640319 38.635774 38.631365
#>  [99] 38.627163 38.628550
#> 
#> $cvse
#>   [1] 12.333497 12.011294 10.648659  9.357182  8.241627  7.278210  6.447861
#>   [8]  5.735682  5.130410  4.623803  4.209894  3.884040  3.641795  3.477781
#>  [15]  3.384906  3.353739  3.243853  3.180456  3.180402  3.193286  3.227767
#>  [22]  3.272643  3.324208  3.379872  3.440638  3.486996  3.517126  3.512890
#>  [29]  3.634415  3.619983  3.593366  3.462105  3.368711  3.322059  3.191504
#>  [36]  3.174131  3.169288  3.169271  3.169264  3.169261  6.746666  9.793869
#>  [43]  9.703346  9.710225  9.700823  9.852388  9.853978  9.848139  9.830048
#>  [50]  9.810219  9.830240  9.853723  9.855245  9.808051  9.856615  9.989817
#>  [57] 10.012720 10.012923  8.446089  8.445235 14.084992 14.096284 14.097402
#>  [64] 13.965388 13.945080 13.879944 13.760274 13.520259 13.521657 13.513128
#>  [71] 13.514501 13.515857 13.517195 12.920635 12.922750 12.923499 12.924251
#>  [78] 12.918268 12.919023 12.921979 12.922736 12.930531 12.932322 12.934445
#>  [85] 12.935214 12.935978 12.936737 12.937491 12.938527 13.055835 13.014497
#>  [92] 13.012097 13.011872 13.009593 13.007371 13.005206 13.003097 13.001041
#>  [99] 12.999028 12.996285
#> 
#> $fold
#>  [1]  3  8  5  7  4  6  5  2 10  9  3  2  9  1  6  8 10  1  7  1  4
#> 
#> $lambda
#>   [1] 6.086301930 5.676105550 5.293555033 4.936787141 4.604064211 4.293765692
#>   [7] 4.004380255 3.734498428 3.482805733 3.248076283 3.029166813 2.825011108
#>  [13] 2.634614814 2.457050593 2.291453607 2.137017303 1.992989489 1.858668668
#>  [19] 1.733400620 1.616575220 1.507623461 1.406014685 1.311254001 1.222879870
#>  [25] 1.140461860 1.063598547 0.991915565 0.925063776 0.862717573 0.804573296
#>  [31] 0.750347749 0.699776822 0.652614206 0.608630193 0.567610555 0.529355503
#>  [37] 0.493678714 0.460406421 0.429376569 0.400438026 0.373449843 0.348280573
#>  [43] 0.324807628 0.302916681 0.282501111 0.263461481 0.245705059 0.229145361
#>  [49] 0.213701730 0.199298949 0.185866867 0.173340062 0.161657522 0.150762346
#>  [55] 0.140601469 0.131125400 0.122287987 0.114046186 0.106359855 0.099191557
#>  [61] 0.092506378 0.086271757 0.080457330 0.075034775 0.069977682 0.065261421
#>  [67] 0.060863019 0.056761056 0.052935550 0.049367871 0.046040642 0.042937657
#>  [73] 0.040043803 0.037344984 0.034828057 0.032480763 0.030291668 0.028250111
#>  [79] 0.026346148 0.024570506 0.022914536 0.021370173 0.019929895 0.018586687
#>  [85] 0.017334006 0.016165752 0.015076235 0.014060147 0.013112540 0.012228799
#>  [91] 0.011404619 0.010635985 0.009919156 0.009250638 0.008627176 0.008045733
#>  [97] 0.007503477 0.006997768 0.006526142 0.006086302
#> 
#> $fit
#> $beta
#>              6.0863       5.6761       5.2936      4.9368      4.6041    4.2938
#> (Intercept) 20.2619 21.369624588 22.402687805 23.36612602 24.26463172 25.102581
#> am           0.0000  0.000000000  0.000000000  0.00000000  0.00000000  0.000000
#> carb         0.0000  0.000000000  0.000000000  0.00000000  0.00000000  0.000000
#> cyl          0.0000  0.000000000  0.000000000  0.00000000  0.00000000  0.000000
#> disp         0.0000 -0.004648797 -0.008984281 -0.01302757 -0.01679835 -0.020315
#> drat         0.0000  0.000000000  0.000000000  0.00000000  0.00000000  0.000000
#> gear         0.0000  0.000000000  0.000000000  0.00000000  0.00000000  0.000000
#> hp           0.0000  0.000000000  0.000000000  0.00000000  0.00000000  0.000000
#> qsec         0.0000  0.000000000  0.000000000  0.00000000  0.00000000  0.000000
#> vs           0.0000  0.000000000  0.000000000  0.00000000  0.00000000  0.000000
#> wt           0.0000  0.000000000  0.000000000  0.00000000  0.00000000  0.000000
#>                  4.0044      3.7345      3.4828     3.2481      3.0292
#> (Intercept) 25.88405559 26.61286126 27.29254788 27.9264259 28.51758264
#> am           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> disp        -0.02359463 -0.02665323 -0.02950569 -0.0321659 -0.03464682
#> drat         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#>                  2.8250      2.6346      2.4571     2.2915      2.1370
#> (Intercept) 29.06889741 29.58305541 30.06256088 30.5097493 30.92679869
#> am           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> disp        -0.03696054 -0.03911832 -0.04113067 -0.0430074 -0.04475764
#> drat         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#>                  1.9930      1.8587      1.7334      1.6166      1.5076
#> (Intercept) 31.21912370 31.21912370 31.21912370 31.21912370 31.21912370
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.04598445 -0.04598445 -0.04598445 -0.04598445 -0.04598445
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  1.4060      1.3113      1.2229      1.1405      1.0636
#> (Intercept) 31.21912370 31.21912370 31.21912370 31.21912370 31.00326994
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.26025371
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.04598445 -0.04598445 -0.04598445 -0.04598445 -0.04549465
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.9919      0.9251      0.8627      0.8046      0.7503
#> (Intercept) 30.70003078 30.41707693 30.15319299 29.63183654 28.90287899
#> am           0.62586750  0.96702341  1.28518684  2.27707828  3.81263758
#> carb         0.00000000  0.00000000  0.00000000 -0.27745469 -0.77895692
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.04480657 -0.04416451 -0.04356573 -0.03980301 -0.03348608
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.6998      0.6526      0.6086      0.5676     0.5294
#> (Intercept) 28.57802089 28.57084551 28.57048806 28.57023782 28.5700893
#> am           4.59106697  4.77193432  4.78164262  4.78216906  4.7824814
#> carb        -1.07432457 -1.20584900 -1.21293426 -1.21310588 -1.2132077
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> disp        -0.03000269 -0.02876353 -0.02869684 -0.02869468 -0.0286934
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#>                  0.4937      0.4604      0.4294      0.4004      0.3734
#> (Intercept) 28.57000123 28.56994895 28.56991793 28.56989953 28.56988861
#> am           4.78266677  4.78277675  4.78284200  4.78288072  4.78290370
#> carb        -1.21326814 -1.21330400 -1.21332527 -1.21333789 -1.21334538
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.02869263 -0.02869218 -0.02869191 -0.02869176 -0.02869166
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.3483      0.3248      0.3029      0.2825      0.2635
#> (Intercept) 28.56988213 28.56987828 28.56987600 28.56987465 28.56987384
#> am           4.78291733  4.78292542  4.78293022  4.78293307  4.78293476
#> carb        -1.21334983 -1.21335246 -1.21335403 -1.21335496 -1.21335551
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.02869161 -0.02869157 -0.02869155 -0.02869154 -0.02869153
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.2457      0.2291      0.2137      0.1993      0.1859
#> (Intercept) 28.56987337 22.92027573 20.90091544 20.89840126 20.89654608
#> am           4.78293576  5.40594040  5.62973968  5.63042051  5.63092305
#> carb        -1.21335583 -1.14681976 -1.12312813 -1.12320825 -1.12326744
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.02869153 -0.02610061 -0.02516987 -0.02516705 -0.02516496
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.25576122  0.34710760  0.34720738  0.34728101
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.1733      0.1617      0.1508      0.1406      0.1311
#> (Intercept) 18.86010388 18.85810844 18.85651366 18.85523896 18.85422004
#> am           5.18599251  5.18576626  5.18558522  5.18544040  5.18532457
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.02962565 -0.02961976 -0.02961505 -0.02961129 -0.02960828
#> qsec         0.66052085  0.66063895  0.66073335  0.66080881  0.66086913
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -2.47385990 -2.47414338 -2.47437006 -2.47455132 -2.47469626
#>                  0.1223      0.1140      0.1064      0.0992     0.0925
#> (Intercept) 18.85340552 18.85275438 18.85223382 18.85181766 18.8514849
#> am           5.18523193  5.18515784  5.18509860  5.18505123  5.1850134
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> hp          -0.02960587 -0.02960395 -0.02960241 -0.02960118 -0.0296002
#> qsec         0.66091736  0.66095591  0.66098673  0.66101138  0.6610311
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> wt          -2.47481214 -2.47490480 -2.47497888 -2.47503811 -2.4750855
#>                  0.0863      0.0805      0.0750      0.0700      0.0653
#> (Intercept) 18.85121895 18.85100628 18.85083626 16.88198644 16.87722665
#> am           5.18498307  5.18495885  5.18493949  5.31495002  5.31507201
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.02959941 -0.02959879 -0.02959828 -0.02913924 -0.02913329
#> qsec         0.66104683  0.66105942  0.66106949  0.80902584  0.80931263
#> vs           0.00000000  0.00000000  0.00000000 -0.74036850 -0.74102759
#> wt          -2.47512333 -2.47515361 -2.47517781 -2.61781308 -2.61812937
#>                  0.0609      0.0568      0.0529      0.0494      0.0460
#> (Intercept) 16.87314004 16.86963108 16.86661809 16.86403108 19.06593437
#> am           5.31517737  5.31526839  5.31534698  5.31541481  5.77037708
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.27350479
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         0.00000000  0.00000000  0.00000000  0.00000000 -0.41693220
#> gear         0.00000000  0.00000000  0.00000000  0.00000000 -0.29640227
#> hp          -0.02912819 -0.02912381 -0.02912006 -0.02911684 -0.03372974
#> qsec         0.80955885  0.80977024  0.80995172  0.81010753  0.86246411
#> vs          -0.74159391 -0.74208037 -0.74249810 -0.74285674 -0.97422387
#> wt          -2.61840051 -2.61863295 -2.61883223 -2.61900311 -2.82093509
#>                  0.0429      0.0400      0.0373      0.0348       0.0325
#> (Intercept) 19.06416408 19.06241734 19.06071162 19.05906169 22.069013344
#> am           5.77118445  5.77191618  5.77257617  5.77316891  5.714520107
#> carb         0.27385172  0.27415485  0.27441903  0.27464861  0.429059043
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000 -0.331472953
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.005995808
#> drat        -0.41738743 -0.41778390 -0.41812838 -0.41842709 -0.591466482
#> gear        -0.29677305 -0.29710591 -0.29740333 -0.29766796 -0.426218826
#> hp          -0.03373158 -0.03373294 -0.03373392 -0.03373456 -0.035443904
#> qsec         0.86278064  0.86306972  0.86333309  0.86357243  0.856910825
#> vs          -0.97518463 -0.97605086 -0.97683029 -0.97753018 -1.031652868
#> wt          -2.82137571 -2.82176985 -2.82212201 -2.82243630 -3.224338217
#>                   0.0303       0.0283      0.0263       0.0246       0.0229
#> (Intercept) 22.071986477 22.074774124 22.07738588 22.079830914 22.082118020
#> am           5.714599371  5.714670936  5.71473538  5.714793253  5.714845058
#> carb         0.429217543  0.429363963  0.42949909  0.429623645  0.429738338
#> cyl         -0.331790784 -0.332087327 -0.33236376 -0.332621217 -0.332860766
#> disp         0.005998152  0.006000257  0.00600214  0.006003818  0.006005307
#> drat        -0.591673913 -0.591864867 -0.59204041 -0.592201569 -0.592349286
#> gear        -0.426505262 -0.426773332 -0.42702403 -0.427258297 -0.427477041
#> hp          -0.035443087 -0.035442283 -0.03544149 -0.035440720 -0.035439966
#> qsec         0.856952945  0.856991033  0.85702539  0.857056289  0.857083998
#> vs          -1.032123233 -1.032565875 -1.03298217 -1.033373426 -1.033740917
#> wt          -3.224613807 -3.224864189 -3.22509115 -3.225296377 -3.225481449
#>                   0.0214       0.0199       0.0186       0.0173       0.0162
#> (Intercept) 22.084255599 22.086251687 22.088113966 22.089849778 22.091466134
#> am           5.714891275  5.714932350  5.714968700  5.715000715  5.715028757
#> carb         0.429843819  0.429940705  0.430029576  0.430110980  0.430185432
#> cyl         -0.333083434 -0.333290196 -0.333481984 -0.333659681 -0.333824129
#> disp         0.006006621  0.006007774  0.006008781  0.006009651  0.006010398
#> drat        -0.592484466 -0.592607962 -0.592720576 -0.592823065 -0.592916140
#> gear        -0.427681123 -0.427871364 -0.428048548 -0.428213420 -0.428366692
#> hp          -0.035439232 -0.035438518 -0.035437827 -0.035437157 -0.035436511
#> qsec         0.857108764  0.857130816  0.857150371  0.857167632  0.857182787
#> vs          -1.034085854 -1.034409396 -1.034712657 -1.034996702 -1.035262549
#> wt          -3.225647855 -3.225796993 -3.225930175 -3.226048632 -3.226153517
#>                   0.0151       0.0141       0.0131       0.0122       0.0114
#> (Intercept) 22.092969731 22.094366958 22.095663911 22.096866401 22.097979966
#> am           5.715053166  5.715074257  5.715092323  5.715107636  5.715120450
#> carb         0.430253416  0.430315386  0.430371771  0.430422971  0.430469365
#> cyl         -0.333976129 -0.334116442 -0.334245790 -0.334364859 -0.334474299
#> disp         0.006011031  0.006011561  0.006011996  0.006012346  0.006012618
#> drat        -0.593000473 -0.593076694 -0.593145395 -0.593207135 -0.593262436
#> gear        -0.428509039 -0.428641105 -0.428763501 -0.428876808 -0.428981580
#> hp          -0.035435888 -0.035435289 -0.035434713 -0.035434161 -0.035433632
#> qsec         0.857196014  0.857207477  0.857217329  0.857225714  0.857232764
#> vs          -1.035511174 -1.035743508 -1.035960440 -1.036162821 -1.036351461
#> wt          -3.226245911 -3.226326824 -3.226397201 -3.226457926 -3.226509824
#>                  0.0106       0.0099       0.0093      0.0086       0.0080
#> (Intercept) 22.09900988 22.099961169 22.100838605 22.10164673 22.102389866
#> am           5.71513100  5.715139501  5.715146157  5.71515115  5.715154660
#> carb         0.43051131  0.430549123  0.430583130  0.43061362  0.430640855
#> cyl         -0.33457473 -0.334666729 -0.334750857 -0.33482764 -0.334897559
#> disp         0.00601282  0.006012958  0.006013039  0.00601307  0.006013055
#> drat        -0.59331179 -0.593355655 -0.593394463 -0.59342862 -0.593458497
#> gear        -0.42907834 -0.429167585 -0.429249788 -0.42932540 -0.429394832
#> hp          -0.03543313 -0.035432644 -0.035432183 -0.03543175 -0.035431328
#> qsec         0.85723860  0.857243344  0.857247095  0.85724995  0.857252010
#> vs          -1.03652714 -1.036690581 -1.036842500 -1.03698356 -1.037114409
#> wt          -3.22655366 -3.226590159 -3.226619979 -3.22664374 -3.226662031
#>                   0.0075       0.0070       0.0065       0.0061
#> (Intercept) 22.103072106 22.103697343 22.104269268 22.104791377
#> am           5.715156837  5.715157831  5.715157775  5.715156792
#> carb         0.430665101  0.430686591  0.430705549  0.430722182
#> cyl         -0.334961097 -0.335018690 -0.335070758 -0.335117692
#> disp         0.006012999  0.006012908  0.006012786  0.006012636
#> drat        -0.593484453 -0.593506816 -0.593525894 -0.593541973
#> gear        -0.429458498 -0.429516773 -0.429570015 -0.429618566
#> hp          -0.035430933 -0.035430558 -0.035430202 -0.035429866
#> qsec         0.857253348  0.857254047  0.857254176  0.857253801
#> vs          -1.037235641 -1.037347839 -1.037451549 -1.037547293
#> wt          -3.226675374 -3.226684268 -3.226689173 -3.226690514
#> 
#> $iter
#>   [1]   0   3   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1   1   1   8   7   7   7  46  45  21  17   5   1   1
#>  [37]   1   1   1   1   1   1   1   1   1   1   1  93  20   1   1 145   1   1
#>  [55]   1   1   1   1   1   1   1   1   1   1  35   1   1   1   1   1  92   1
#>  [73]   1   1   1  83   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 6.086301930 5.676105550 5.293555033 4.936787141 4.604064211 4.293765692
#>   [7] 4.004380255 3.734498428 3.482805733 3.248076283 3.029166813 2.825011108
#>  [13] 2.634614814 2.457050593 2.291453607 2.137017303 1.992989489 1.858668668
#>  [19] 1.733400620 1.616575220 1.507623461 1.406014685 1.311254001 1.222879870
#>  [25] 1.140461860 1.063598547 0.991915565 0.925063776 0.862717573 0.804573296
#>  [31] 0.750347749 0.699776822 0.652614206 0.608630193 0.567610555 0.529355503
#>  [37] 0.493678714 0.460406421 0.429376569 0.400438026 0.373449843 0.348280573
#>  [43] 0.324807628 0.302916681 0.282501111 0.263461481 0.245705059 0.229145361
#>  [49] 0.213701730 0.199298949 0.185866867 0.173340062 0.161657522 0.150762346
#>  [55] 0.140601469 0.131125400 0.122287987 0.114046186 0.106359855 0.099191557
#>  [61] 0.092506378 0.086271757 0.080457330 0.075034775 0.069977682 0.065261421
#>  [67] 0.060863019 0.056761056 0.052935550 0.049367871 0.046040642 0.042937657
#>  [73] 0.040043803 0.037344984 0.034828057 0.032480763 0.030291668 0.028250111
#>  [79] 0.026346148 0.024570506 0.022914536 0.021370173 0.019929895 0.018586687
#>  [85] 0.017334006 0.016165752 0.015076235 0.014060147 0.013112540 0.012228799
#>  [91] 0.011404619 0.010635985 0.009919156 0.009250638 0.008627176 0.008045733
#>  [97] 0.007503477 0.006997768 0.006526142 0.006086302
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 29
#> 
#> $loss
#>   [1] 983.1895 833.8554 708.9152 604.8584 518.6543 447.6878 389.7040 342.7598
#>   [9] 305.1823 275.5323 252.5728 235.2417 222.6282 213.9519 208.5455 205.8387
#>  [17] 205.2850 205.2850 205.2850 205.2850 205.2850 205.2850 205.2850 205.2850
#>  [25] 205.2850 199.6153 192.4854 186.7123 182.0939 157.2861 131.6081 125.9832
#>  [33] 125.4010 125.3993 125.3993 125.3993 125.3992 125.3992 125.3992 125.3992
#>  [41] 125.3992 125.3992 125.3992 125.3992 125.3992 125.3992 125.3992 122.1906
#>  [49] 121.9502 121.9502 121.9502 103.8009 103.8009 103.8009 103.8009 103.8009
#>  [57] 103.8009 103.8009 103.8009 103.8009 103.8009 103.8009 103.8009 103.8009
#>  [65] 103.2182 103.2182 103.2182 103.2181 103.2181 103.2181 102.5022 102.5022
#>  [73] 102.5021 102.5021 102.5021 101.6816 101.6816 101.6816 101.6816 101.6816
#>  [81] 101.6816 101.6816 101.6815 101.6815 101.6815 101.6815 101.6815 101.6815
#>  [89] 101.6815 101.6815 101.6815 101.6815 101.6815 101.6815 101.6815 101.6815
#>  [97] 101.6815 101.6815 101.6815 101.6815
#> 
#> $linear.predictors
#>     6.0863   5.6761   5.2936   4.9368   4.6041   4.2938   4.0044   3.7345
#> 1  20.2619 19.69606 19.16835 18.67620 18.21723 17.78918 17.38999 17.01770
#> 2  20.2619 20.32365 20.38122 20.43492 20.48500 20.53171 20.57526 20.61589
#> 3  20.2619 19.69606 19.16835 18.67620 18.21723 17.78918 17.38999 17.01770
#> 4  20.2619 20.68765 21.08469 21.45498 21.80031 22.12237 22.42272 22.70283
#> 5  20.2619 20.71507 21.13770 21.53184 21.89942 22.24223 22.56193 22.86009
#> 6  20.2619 20.59049 20.89692 21.18271 21.44923 21.69779 21.92960 22.14578
#> 7  20.2619 20.08749 19.92482 19.77312 19.63165 19.49971 19.37666 19.26190
#> 8  20.2619 19.17539 18.16211 17.21711 16.33581 15.51390 14.74739 14.03254
#> 9  20.2619 19.23118 18.26992 17.37344 16.53739 15.75768 15.03053 14.35238
#> 10 20.2619 19.32415 18.44960 17.63400 16.87336 16.16398 15.50242 14.88544
#> 11 20.2619 21.00376 21.69562 22.34086 22.94260 23.50379 24.02716 24.51525
#> 12 20.2619 21.01771 21.72258 22.37994 22.99300 23.56474 24.09794 24.59521
#> 13 20.2619 21.03910 21.76391 22.43987 23.07027 23.65818 24.20648 24.71782
#> 14 20.2619 20.81130 21.32368 21.80152 22.24715 22.66275 23.05034 23.41181
#> 15 20.2619 19.89131 19.54569 19.22336 18.92276 18.64241 18.38096 18.13713
#> 16 20.2619 19.95639 19.67147 19.40575 19.15793 18.92682 18.71129 18.51028
#> 17 20.2619 20.81037 21.32188 21.79891 22.24379 22.65869 23.04562 23.40648
#> 18 20.2619 20.92752 21.54828 22.12720 22.66711 23.17063 23.64021 24.07814
#> 19 20.2619 19.73790 19.24921 18.79345 18.36841 17.97202 17.60234 17.25758
#> 20 20.2619 19.97034 19.69842 19.44483 19.20833 18.98777 18.78207 18.59024
#> 21 20.2619 20.80712 21.31559 21.78979 22.23203 22.64447 23.02911 23.38782
#>      3.4828   3.2481   3.0292   2.8250   2.6346   2.4571   2.2915    2.1370
#> 1  16.67050 16.34670 16.04473 15.76310 15.50046 15.25552 15.02708 14.814047
#> 2  20.65377 20.68910 20.72205 20.75278 20.78143 20.80816 20.83308 20.856329
#> 3  16.67050 16.34670 16.04473 15.76310 15.50046 15.25552 15.02708 14.814047
#> 4  22.96406 23.20769 23.43489 23.64679 23.84440 24.02869 24.20056 24.360852
#> 5  23.13815 23.39747 23.63931 23.86485 24.07520 24.27136 24.45431 24.624922
#> 6  22.34739 22.53542 22.71078 22.87431 23.02682 23.16906 23.30171 23.425418
#> 7  19.15488 19.05507 18.96199 18.87518 18.79422 18.71872 18.64831 18.582641
#> 8  13.36586 12.74412 12.16428 11.62352 11.11921 10.64888 10.21026  9.801191
#> 9  13.71993 13.13011 12.58004 12.06705 11.58863 11.14245 10.72634 10.338283
#> 10 14.31005 13.77343 13.27298 12.80626 12.37099 11.96506 11.58649 11.233436
#> 11 24.97045 25.39497 25.79088 26.16010 26.50444 26.82558 27.12507 27.404372
#> 12 25.05897 25.49147 25.89482 26.27098 26.62180 26.94897 27.25409 27.538645
#> 13 25.19469 25.63943 26.05419 26.44100 26.80174 27.13817 27.45192 27.744530
#> 14 23.74891 24.06330 24.35650 24.62994 24.88495 25.12277 25.34456 25.551406
#> 15 17.90974 17.69767 17.49989 17.31545 17.14343 16.98301 16.83340 16.693868
#> 16 18.32282 18.14799 17.98495 17.83289 17.69109 17.55884 17.43550 17.320475
#> 17 23.74301 24.05687 24.34957 24.62254 24.87712 25.11454 25.33596 25.542454
#> 18 24.48656 24.86745 25.22267 25.55395 25.86290 26.15103 26.41975 26.670347
#> 19 16.93605 16.63620 16.35655 16.09575 15.85252 15.62569 15.41415 15.216866
#> 20 18.41134 18.24449 18.08889 17.94377 17.80844 17.68223 17.56452 17.454748
#> 21 23.72236 24.03435 24.32532 24.59667 24.84974 25.08575 25.30585 25.511124
#>       1.9930    1.8587    1.7334    1.6166    1.5076    1.4060    1.3113
#> 1  14.664721 14.664721 14.664721 14.664721 14.664721 14.664721 14.664721
#> 2  20.872622 20.872622 20.872622 20.872622 20.872622 20.872622 20.872622
#> 3  14.664721 14.664721 14.664721 14.664721 14.664721 14.664721 14.664721
#> 4  24.473205 24.473205 24.473205 24.473205 24.473205 24.473205 24.473205
#> 5  24.744513 24.744513 24.744513 24.744513 24.744513 24.744513 24.744513
#> 6  23.512130 23.512130 23.512130 23.512130 23.512130 23.512130 23.512130
#> 7  18.536612 18.536612 18.536612 18.536612 18.536612 18.536612 18.536612
#> 8   9.514463  9.514463  9.514463  9.514463  9.514463  9.514463  9.514463
#> 9  10.066276 10.066276 10.066276 10.066276 10.066276 10.066276 10.066276
#> 10 10.985965 10.985965 10.985965 10.985965 10.985965 10.985965 10.985965
#> 11 27.600147 27.600147 27.600147 27.600147 27.600147 27.600147 27.600147
#> 12 27.738101 27.738101 27.738101 27.738101 27.738101 27.738101 27.738101
#> 13 27.949629 27.949629 27.949629 27.949629 27.949629 27.949629 27.949629
#> 14 25.696391 25.696391 25.696391 25.696391 25.696391 25.696391 25.696391
#> 15 16.596068 16.596068 16.596068 16.596068 16.596068 16.596068 16.596068
#> 16 17.239850 17.239850 17.239850 17.239850 17.239850 17.239850 17.239850
#> 17 25.687194 25.687194 25.687194 25.687194 25.687194 25.687194 25.687194
#> 18 26.846002 26.846002 26.846002 26.846002 26.846002 26.846002 26.846002
#> 19 15.078581 15.078581 15.078581 15.078581 15.078581 15.078581 15.078581
#> 20 17.377804 17.377804 17.377804 17.377804 17.377804 17.377804 17.377804
#> 21 25.655005 25.655005 25.655005 25.655005 25.655005 25.655005 25.655005
#>       1.2229    1.1405    1.0636    0.9919    0.9251    0.8627    0.8046
#> 1  14.664721 14.664721 14.625195 14.569666 14.517853 14.469531 14.747844
#> 2  20.872622 20.872622 20.766973 20.618553 20.480062 20.350904 20.398705
#> 3  14.664721 14.664721 14.625195 14.569666 14.517853 14.469531 14.192935
#> 4  24.473205 24.473205 24.329204 24.126907 23.938143 23.762101 23.237826
#> 5  24.744513 24.744513 24.597623 24.391266 24.198714 24.019138 23.472664
#> 6  23.512130 23.512130 23.378366 23.190450 23.015105 22.851577 21.851033
#> 7  18.536612 18.536612 18.455844 18.342379 18.236504 18.137765 17.821803
#> 8   9.514463  9.514463  9.529793  9.551331  9.571427  9.590169  9.734998
#> 9  10.066276 10.066276 10.075729 10.089009 10.101401 10.112958 10.212634
#> 10 10.985965 10.985965 10.985622 10.985141 10.984692 10.984273 11.008694
#> 11 27.600147 27.600147 27.683094 27.799621 27.908353 28.009757 28.498963
#> 12 27.738101 27.738101 27.819578 27.934041 28.040847 28.140454 28.340918
#> 13 27.949629 27.949629 28.028854 28.140151 28.244004 28.340857 28.801466
#> 14 25.696391 25.696391 25.539362 25.318762 25.112919 24.920949 24.574040
#> 15 16.596068 16.596068 16.535970 16.451542 16.372762 16.299291 16.419570
#> 16 17.239850 17.239850 17.172895 17.078834 16.991065 16.909212 16.976812
#> 17 25.687194 25.687194 25.790517 25.935668 26.071110 26.197423 26.565703
#> 18 26.846002 26.846002 26.936982 27.064794 27.184055 27.295279 27.568739
#> 19 15.078581 15.078581 15.294900 15.598793 15.882357 16.146809 16.828240
#> 20 17.377804 17.377804 17.569633 17.839121 18.090582 18.325096 17.708572
#> 21 25.655005 25.655005 25.758671 25.904304 26.040194 26.166927 26.537841
#>      0.7503   0.6998   0.6526   0.6086   0.5676   0.5294   0.4937   0.4604
#> 1  15.28997 15.62840 15.80428 15.81376 15.81394 15.81405 15.81412 15.81416
#> 2  20.58955 20.75309 20.89320 20.90076 20.90083 20.90087 20.90089 20.90090
#> 3  13.73206 13.47975 13.39258 13.38789 13.38773 13.38764 13.38758 13.38755
#> 4  22.43256 22.02798 21.93954 21.93479 21.93452 21.93435 21.93426 21.93420
#> 5  22.63012 22.20499 22.10924 22.10410 22.10382 22.10364 22.10354 22.10348
#> 6  20.17478 19.25227 18.92668 18.90916 18.90859 18.90825 18.90804 18.90792
#> 7  17.33055 17.08030 17.02032 17.01710 17.01693 17.01683 17.01677 17.01673
#> 8   9.98162 10.11945 10.17106 10.17384 10.17393 10.17398 10.17400 10.17402
#> 9  10.38345 10.47948 10.51622 10.51820 10.51826 10.51830 10.51832 10.51833
#> 10 11.05317 11.07954 11.09150 11.09214 11.09215 11.09216 11.09217 11.09217
#> 11 29.30120 29.73355 29.87324 29.88075 29.88103 29.88119 29.88129 29.88135
#> 12 28.62271 28.74923 28.75368 28.75391 28.75401 28.75407 28.75410 28.75412
#> 13 29.55570 29.96157 30.09184 30.09885 30.09911 30.09926 30.09935 30.09941
#> 14 24.10224 23.90037 23.91050 23.91106 23.91090 23.91080 23.91075 23.91071
#> 15 16.69639 16.88852 17.01234 17.01902 17.01912 17.01917 17.01921 17.01923
#> 16 17.16520 17.30855 17.41503 17.42078 17.42084 17.42088 17.42090 17.42092
#> 17 27.12923 27.41111 27.47083 27.47403 27.47423 27.47434 27.47441 27.47445
#> 18 27.97308 28.16718 28.19567 28.19719 28.19733 28.19741 28.19746 28.19749
#> 19 17.84607 18.34084 18.42338 18.42780 18.42815 18.42836 18.42848 18.42855
#> 20 16.40455 15.54368 15.03816 15.01091 15.01046 15.01020 15.01004 15.00995
#> 21 27.10579 27.39011 27.45069 27.45394 27.45414 27.45425 27.45432 27.45436
#>      0.4294   0.4004   0.3734   0.3483   0.3248   0.3029   0.2825   0.2635
#> 1  15.81418 15.81419 15.81420 15.81420 15.81421 15.81421 15.81421 15.81421
#> 2  20.90091 20.90092 20.90092 20.90092 20.90092 20.90092 20.90092 20.90092
#> 3  13.38753 13.38752 13.38751 13.38750 13.38750 13.38750 13.38750 13.38750
#> 4  21.93416 21.93414 21.93413 21.93412 21.93412 21.93412 21.93412 21.93411
#> 5  22.10345 22.10342 22.10341 22.10340 22.10340 22.10340 22.10340 22.10339
#> 6  18.90785 18.90781 18.90778 18.90777 18.90776 18.90776 18.90775 18.90775
#> 7  17.01671 17.01670 17.01669 17.01669 17.01669 17.01668 17.01668 17.01668
#> 8  10.17403 10.17404 10.17404 10.17405 10.17405 10.17405 10.17405 10.17405
#> 9  10.51834 10.51834 10.51834 10.51834 10.51835 10.51835 10.51835 10.51835
#> 10 11.09217 11.09218 11.09218 11.09218 11.09218 11.09218 11.09218 11.09218
#> 11 29.88138 29.88140 29.88141 29.88142 29.88142 29.88143 29.88143 29.88143
#> 12 28.75413 28.75414 28.75414 28.75415 28.75415 28.75415 28.75415 28.75415
#> 13 30.09944 30.09946 30.09947 30.09948 30.09948 30.09948 30.09948 30.09949
#> 14 23.91069 23.91068 23.91067 23.91067 23.91067 23.91067 23.91067 23.91067
#> 15 17.01924 17.01925 17.01925 17.01925 17.01925 17.01925 17.01925 17.01926
#> 16 17.42093 17.42093 17.42093 17.42093 17.42094 17.42094 17.42094 17.42094
#> 17 27.47447 27.47449 27.47449 27.47450 27.47450 27.47450 27.47451 27.47451
#> 18 28.19751 28.19752 28.19752 28.19753 28.19753 28.19753 28.19753 28.19753
#> 19 18.42860 18.42862 18.42864 18.42865 18.42865 18.42866 18.42866 18.42866
#> 20 15.00989 15.00986 15.00984 15.00983 15.00982 15.00982 15.00981 15.00981
#> 21 27.45439 27.45440 27.45441 27.45442 27.45442 27.45442 27.45442 27.45442
#>      0.2457   0.2291   0.2137   0.1993   0.1859   0.1733   0.1617   0.1508
#> 1  15.81421 15.58347 15.50128 15.50132 15.50135 16.40760 16.40767 16.40773
#> 2  20.90092 21.07231 21.13308 21.13314 21.13318 20.54559 20.54562 20.54564
#> 3  13.38750 12.98803 12.84544 12.84520 12.84502 13.23279 13.23310 13.23334
#> 4  21.93411 21.91290 21.90439 21.90413 21.90393 22.34212 22.34195 22.34181
#> 5  22.10339 22.80860 23.05951 23.05951 23.05952 23.37894 23.37931 23.37961
#> 6  18.90775 18.79242 18.75027 18.74979 18.74944 19.18992 19.18990 19.18989
#> 7  17.01668 16.73151 16.62935 16.62911 16.62894 14.95194 14.95191 14.95188
#> 8  10.17405 10.61210 10.76922 10.76951 10.76973 11.67525 11.67509 11.67497
#> 9  10.51835 10.88438 11.01572 11.01596 11.01614 10.84286 10.84269 10.84256
#> 10 11.09218 11.30409 11.38028 11.38042 11.38053 10.32970 10.32960 10.32952
#> 11 29.88143 30.10495 30.18484 30.18509 30.18528 29.50865 29.50850 29.50837
#> 12 28.75415 28.79346 28.80747 28.80754 28.80759 30.74313 30.74294 30.74279
#> 13 30.09949 30.41329 30.52539 30.52566 30.52586 30.72526 30.72525 30.72525
#> 14 23.91066 23.75655 23.70051 23.70025 23.70006 23.10537 23.10561 23.10581
#> 15 17.01926 16.64133 16.50635 16.50625 16.50618 16.85126 16.85114 16.85104
#> 16 17.42094 17.11672 17.00798 17.00789 17.00782 17.34556 17.34552 17.34548
#> 17 27.47451 27.16389 27.05316 27.05317 27.05318 27.08680 27.08648 27.08622
#> 18 28.19753 27.87277 27.75686 27.75682 27.75680 28.11825 28.11826 28.11827
#> 19 18.42866 18.28616 18.23658 18.23686 18.23707 17.96034 17.96049 17.96061
#> 20 15.00981 15.02949 15.03727 15.03710 15.03698 14.93343 14.93389 14.93426
#> 21 27.45442 27.63156 27.69505 27.69525 27.69540 26.22526 26.22509 26.22495
#>      0.1406   0.1311   0.1223   0.1140   0.1064   0.0992   0.0925   0.0863
#> 1  16.40777 16.40781 16.40784 16.40786 16.40788 16.40789 16.40790 16.40791
#> 2  20.54566 20.54568 20.54569 20.54570 20.54571 20.54571 20.54572 20.54572
#> 3  13.23354 13.23369 13.23382 13.23392 13.23400 13.23406 13.23411 13.23415
#> 4  22.34170 22.34161 22.34154 22.34148 22.34144 22.34140 22.34137 22.34135
#> 5  23.37985 23.38004 23.38020 23.38032 23.38042 23.38050 23.38056 23.38061
#> 6  19.18988 19.18987 19.18987 19.18986 19.18986 19.18986 19.18985 19.18985
#> 7  14.95186 14.95184 14.95183 14.95181 14.95180 14.95180 14.95179 14.95179
#> 8  11.67487 11.67479 11.67473 11.67468 11.67464 11.67461 11.67458 11.67456
#> 9  10.84246 10.84238 10.84231 10.84226 10.84221 10.84218 10.84215 10.84213
#> 10 10.32946 10.32940 10.32936 10.32933 10.32931 10.32929 10.32927 10.32926
#> 11 29.50827 29.50819 29.50812 29.50807 29.50803 29.50800 29.50797 29.50795
#> 12 30.74267 30.74258 30.74250 30.74244 30.74239 30.74235 30.74232 30.74230
#> 13 30.72524 30.72523 30.72523 30.72523 30.72523 30.72522 30.72522 30.72522
#> 14 23.10596 23.10608 23.10618 23.10626 23.10632 23.10637 23.10641 23.10644
#> 15 16.85097 16.85091 16.85086 16.85082 16.85079 16.85077 16.85075 16.85073
#> 16 17.34545 17.34543 17.34542 17.34540 17.34539 17.34538 17.34537 17.34537
#> 17 27.08602 27.08586 27.08572 27.08562 27.08554 27.08547 27.08542 27.08537
#> 18 28.11828 28.11828 28.11829 28.11829 28.11829 28.11829 28.11830 28.11830
#> 19 17.96070 17.96077 17.96083 17.96088 17.96092 17.96095 17.96098 17.96099
#> 20 14.93456 14.93480 14.93498 14.93514 14.93526 14.93535 14.93543 14.93549
#> 21 26.22484 26.22475 26.22468 26.22463 26.22458 26.22455 26.22452 26.22449
#>      0.0805   0.0750   0.0700   0.0653   0.0609   0.0568   0.0529   0.0494
#> 1  16.40792 16.40793 16.54696 16.54704 16.54710 16.54716 16.54720 16.54724
#> 2  20.54572 20.54573 20.38287 20.38278 20.38270 20.38263 20.38258 20.38253
#> 3  13.23419 13.23421 13.21225 13.21236 13.21246 13.21254 13.21261 13.21267
#> 4  22.34133 22.34132 22.16468 22.16436 22.16408 22.16384 22.16364 22.16346
#> 5  23.38065 23.38068 23.65397 23.65469 23.65530 23.65583 23.65629 23.65668
#> 6  19.18985 19.18985 18.84280 18.84245 18.84214 18.84188 18.84166 18.84146
#> 7  14.95178 14.95178 15.05947 15.05949 15.05950 15.05951 15.05952 15.05953
#> 8  11.67455 11.67453 11.71121 11.71116 11.71113 11.71110 11.71107 11.71105
#> 9  10.84211 10.84210 10.83487 10.83479 10.83471 10.83465 10.83460 10.83455
#> 10 10.32924 10.32924 10.28098 10.28090 10.28082 10.28076 10.28070 10.28066
#> 11 29.50793 29.50792 29.52592 29.52591 29.52589 29.52588 29.52587 29.52586
#> 12 30.74228 30.74226 30.69672 30.69653 30.69637 30.69623 30.69612 30.69601
#> 13 30.72522 30.72522 30.85844 30.85866 30.85885 30.85901 30.85915 30.85926
#> 14 23.10647 23.10649 23.05081 23.05093 23.05103 23.05111 23.05119 23.05125
#> 15 16.85072 16.85071 16.94466 16.94452 16.94440 16.94430 16.94420 16.94413
#> 16 17.34536 17.34536 17.51506 17.51507 17.51507 17.51508 17.51508 17.51509
#> 17 27.08534 27.08531 27.45388 27.45389 27.45391 27.45392 27.45393 27.45394
#> 18 28.11830 28.11830 27.87562 27.87536 27.87514 27.87495 27.87479 27.87465
#> 19 17.96101 17.96102 17.93659 17.93667 17.93675 17.93682 17.93687 17.93692
#> 20 14.93554 14.93558 14.90148 14.90189 14.90224 14.90255 14.90281 14.90303
#> 21 26.22448 26.22446 26.05075 26.05056 26.05039 26.05025 26.05013 26.05002
#>      0.0460   0.0429   0.0400   0.0373   0.0348   0.0325   0.0303   0.0283
#> 1  16.48282 16.48275 16.48268 16.48263 16.48258 16.58228 16.58227 16.58226
#> 2  20.46224 20.46217 20.46211 20.46205 20.46199 20.36445 20.36429 20.36413
#> 3  13.25930 13.25934 13.25937 13.25940 13.25942 13.49352 13.49378 13.49403
#> 4  22.07389 22.07350 22.07314 22.07280 22.07249 22.21684 22.21684 22.21683
#> 5  23.47889 23.47928 23.47963 23.47996 23.48027 23.48979 23.48989 23.48998
#> 6  18.81354 18.81317 18.81283 18.81250 18.81221 18.49052 18.49014 18.48978
#> 7  15.17158 15.17172 15.17185 15.17196 15.17206 14.67088 14.67070 14.67053
#> 8  11.83173 11.83191 11.83206 11.83220 11.83232 12.16531 12.16550 12.16566
#> 9  10.83641 10.83641 10.83641 10.83641 10.83641 10.99938 10.99948 10.99956
#> 10 10.11244 10.11221 10.11201 10.11184 10.11169 10.12373 10.12374 10.12376
#> 11 29.60886 29.60901 29.60916 29.60929 29.60941 29.46007 29.46000 29.45994
#> 12 30.83109 30.83119 30.83127 30.83135 30.83141 30.93678 30.93680 30.93681
#> 13 30.98472 30.98511 30.98546 30.98579 30.98609 30.91250 30.91250 30.91250
#> 14 22.96588 22.96576 22.96567 22.96559 22.96552 23.15427 23.15456 23.15483
#> 15 17.13362 17.13369 17.13374 17.13378 17.13380 17.06074 17.06066 17.06059
#> 16 17.58166 17.58172 17.58178 17.58182 17.58186 17.38867 17.38852 17.38838
#> 17 27.35124 27.35128 27.35131 27.35134 27.35136 27.47070 27.47091 27.47112
#> 18 27.85136 27.85103 27.85074 27.85046 27.85021 28.09159 28.09162 28.09164
#> 19 17.34758 17.34694 17.34638 17.34589 17.34547 17.17230 17.17191 17.17154
#> 20 15.28817 15.28895 15.28965 15.29027 15.29083 15.27038 15.27060 15.27081
#> 21 26.03299 26.03287 26.03276 26.03267 26.03260 25.98529 25.98531 25.98533
#>      0.0263   0.0246   0.0229   0.0214   0.0199   0.0186   0.0173   0.0162
#> 1  16.58224 16.58223 16.58221 16.58219 16.58217 16.58214 16.58212 16.58209
#> 2  20.36398 20.36384 20.36371 20.36358 20.36346 20.36335 20.36325 20.36315
#> 3  13.49426 13.49448 13.49467 13.49486 13.49503 13.49518 13.49533 13.49546
#> 4  22.21683 22.21683 22.21683 22.21683 22.21683 22.21683 22.21684 22.21684
#> 5  23.49006 23.49013 23.49020 23.49027 23.49032 23.49038 23.49043 23.49048
#> 6  18.48944 18.48914 18.48885 18.48859 18.48835 18.48813 18.48793 18.48774
#> 7  14.67038 14.67024 14.67013 14.67002 14.66993 14.66984 14.66977 14.66971
#> 8  12.16581 12.16595 12.16608 12.16619 12.16629 12.16638 12.16646 12.16653
#> 9  10.99964 10.99971 10.99978 10.99984 10.99990 10.99995 11.00000 11.00004
#> 10 10.12378 10.12380 10.12381 10.12383 10.12385 10.12387 10.12389 10.12391
#> 11 29.45988 29.45982 29.45977 29.45972 29.45968 29.45964 29.45961 29.45958
#> 12 30.93682 30.93683 30.93683 30.93684 30.93684 30.93684 30.93684 30.93684
#> 13 30.91249 30.91249 30.91249 30.91248 30.91248 30.91247 30.91247 30.91246
#> 14 23.15509 23.15532 23.15555 23.15576 23.15595 23.15614 23.15631 23.15647
#> 15 17.06052 17.06046 17.06040 17.06034 17.06029 17.06023 17.06018 17.06014
#> 16 17.38825 17.38813 17.38802 17.38792 17.38782 17.38773 17.38764 17.38757
#> 17 27.47132 27.47150 27.47168 27.47185 27.47200 27.47215 27.47230 27.47243
#> 18 28.09165 28.09166 28.09167 28.09167 28.09167 28.09167 28.09167 28.09166
#> 19 17.17120 17.17088 17.17058 17.17030 17.17004 17.16980 17.16957 17.16936
#> 20 15.27100 15.27118 15.27135 15.27150 15.27165 15.27179 15.27191 15.27203
#> 21 25.98535 25.98538 25.98540 25.98543 25.98545 25.98548 25.98550 25.98553
#>      0.0151   0.0141   0.0131   0.0122   0.0114   0.0106   0.0099   0.0093
#> 1  16.58207 16.58204 16.58202 16.58199 16.58197 16.58194 16.58192 16.58189
#> 2  20.36305 20.36297 20.36288 20.36281 20.36273 20.36267 20.36260 20.36254
#> 3  13.49558 13.49569 13.49580 13.49589 13.49598 13.49605 13.49612 13.49619
#> 4  22.21684 22.21685 22.21685 22.21686 22.21686 22.21687 22.21687 22.21688
#> 5  23.49052 23.49056 23.49060 23.49063 23.49066 23.49069 23.49071 23.49074
#> 6  18.48757 18.48742 18.48727 18.48715 18.48703 18.48692 18.48683 18.48674
#> 7  14.66965 14.66961 14.66957 14.66954 14.66951 14.66949 14.66948 14.66947
#> 8  12.16660 12.16666 12.16671 12.16675 12.16679 12.16682 12.16685 12.16687
#> 9  11.00008 11.00011 11.00015 11.00018 11.00020 11.00023 11.00025 11.00027
#> 10 10.12393 10.12394 10.12396 10.12398 10.12400 10.12401 10.12403 10.12405
#> 11 29.45955 29.45952 29.45950 29.45948 29.45946 29.45944 29.45943 29.45941
#> 12 30.93684 30.93683 30.93683 30.93683 30.93682 30.93682 30.93681 30.93681
#> 13 30.91245 30.91245 30.91244 30.91243 30.91242 30.91242 30.91241 30.91240
#> 14 23.15662 23.15676 23.15689 23.15701 23.15712 23.15722 23.15732 23.15741
#> 15 17.06009 17.06005 17.06002 17.05998 17.05995 17.05991 17.05988 17.05986
#> 16 17.38750 17.38743 17.38737 17.38731 17.38726 17.38721 17.38717 17.38713
#> 17 27.47256 27.47268 27.47279 27.47289 27.47300 27.47309 27.47318 27.47326
#> 18 28.09165 28.09164 28.09163 28.09162 28.09161 28.09159 28.09158 28.09156
#> 19 17.16916 17.16898 17.16880 17.16865 17.16850 17.16836 17.16823 17.16811
#> 20 15.27214 15.27224 15.27233 15.27241 15.27249 15.27257 15.27263 15.27270
#> 21 25.98555 25.98558 25.98560 25.98562 25.98565 25.98567 25.98569 25.98572
#>      0.0086   0.0080   0.0075   0.0070   0.0065   0.0061
#> 1  16.58187 16.58184 16.58182 16.58179 16.58177 16.58175
#> 2  20.36249 20.36243 20.36239 20.36234 20.36230 20.36226
#> 3  13.49625 13.49630 13.49635 13.49639 13.49643 13.49646
#> 4  22.21688 22.21689 22.21689 22.21690 22.21690 22.21691
#> 5  23.49076 23.49078 23.49080 23.49081 23.49082 23.49084
#> 6  18.48667 18.48660 18.48654 18.48648 18.48643 18.48639
#> 7  14.66946 14.66946 14.66946 14.66947 14.66947 14.66948
#> 8  12.16689 12.16691 12.16692 12.16693 12.16694 12.16694
#> 9  11.00028 11.00030 11.00031 11.00032 11.00033 11.00034
#> 10 10.12406 10.12408 10.12409 10.12411 10.12412 10.12413
#> 11 29.45940 29.45939 29.45938 29.45937 29.45937 29.45936
#> 12 30.93680 30.93680 30.93679 30.93679 30.93678 30.93678
#> 13 30.91240 30.91239 30.91238 30.91237 30.91237 30.91236
#> 14 23.15749 23.15756 23.15763 23.15770 23.15776 23.15781
#> 15 17.05983 17.05980 17.05978 17.05976 17.05974 17.05972
#> 16 17.38709 17.38706 17.38703 17.38700 17.38697 17.38695
#> 17 27.47334 27.47341 27.47348 27.47354 27.47360 27.47366
#> 18 28.09155 28.09153 28.09151 28.09150 28.09148 28.09146
#> 19 17.16801 17.16791 17.16781 17.16773 17.16765 17.16758
#> 20 15.27275 15.27280 15.27285 15.27290 15.27293 15.27297
#> 21 25.98574 25.98576 25.98578 25.98580 25.98582 25.98584
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 18.7 18.1 14.3 24.4 22.8 17.8 16.4 10.4 10.4 14.7 32.4 30.4 33.9 21.5 15.5
#> [16] 15.2 26.0 30.4 15.8 15.0 21.4
#> 
#> $X
#>               am       carb         cyl       disp       drat       gear
#>  [1,] -0.7844645 -0.4442617  1.00250941  0.9196362 -0.7858427 -0.9338592
#>  [2,] -0.7844645 -1.0662280 -0.05012547 -0.1003429 -1.4534829 -0.9338592
#>  [3,] -0.7844645  0.7996710  1.00250941  0.9196362 -0.6831288 -0.9338592
#>  [4,] -0.7844645 -0.4442617 -1.10276036 -0.6919308  0.1385822  0.3735437
#>  [5,] -0.7844645 -0.4442617 -1.10276036 -0.7365077  0.5323188  0.3735437
#>  [6,] -0.7844645  0.7996710 -0.05012547 -0.5340229  0.5323188  0.3735437
#>  [7,] -0.7844645  0.1777047  1.00250941  0.2834714 -0.9227945 -0.9338592
#>  [8,] -0.7844645  0.7996710  1.00250941  1.7658411 -1.1624603 -0.9338592
#>  [9,] -0.7844645  0.7996710  1.00250941  1.6751763 -1.0426274 -0.9338592
#> [10,] -0.7844645  0.7996710  1.00250941  1.5240683 -0.6488909 -0.9338592
#> [11,]  1.2747549 -1.0662280 -1.10276036 -1.2056981  0.8062224  0.3735437
#> [12,]  1.2747549 -0.4442617 -1.10276036 -1.2283643  2.2613357  0.3735437
#> [13,]  1.2747549 -1.0662280 -1.10276036 -1.2631191  1.0458882  0.3735437
#> [14,] -0.7844645 -1.0662280 -1.10276036 -0.8929045  0.1557012 -0.9338592
#> [15,] -0.7844645 -0.4442617  1.00250941  0.6023094 -1.4534829 -0.9338592
#> [16,] -0.7844645 -0.4442617  1.00250941  0.4965338 -0.7858427 -0.9338592
#> [17,]  1.2747549 -0.4442617 -1.10276036 -0.8913934  1.4053867  1.6809466
#> [18,]  1.2747549 -0.4442617 -1.10276036 -1.0817895  0.2755341  1.6809466
#> [19,]  1.2747549  0.7996710  1.00250941  0.8516376  1.0458882  1.6809466
#> [20,]  1.2747549  3.2875363  1.00250941  0.4738676 -0.1182025  1.6809466
#> [21,]  1.2747549 -0.4442617 -1.10276036 -0.8861046  0.8575794  0.3735437
#>                hp        qsec         vs          wt
#>  [1,]  0.34790881 -0.52667534 -0.9534626  0.15934567
#>  [2,] -0.58535022  1.12332852  1.0488088  0.17764124
#>  [3,]  1.28116784 -1.13511426 -0.9534626  0.27826685
#>  [4,] -1.15863792  1.00989075  1.0488088 -0.06934891
#>  [5,] -0.71867294  2.50520674  1.0488088 -0.10594004
#>  [6,] -0.34536933  0.44270193  1.0488088  0.15934567
#>  [7,]  0.41457017 -0.33073738 -0.9534626  0.73565601
#>  [8,]  0.74787697 -0.03167418 -0.9534626  1.81509443
#>  [9,]  0.88119969 -0.11417437 -0.9534626  1.97426586
#> [10,]  1.08118376 -0.32042486 -0.9534626  1.90199837
#> [11,] -1.10530883  0.73660886  1.0488088 -0.97497945
#> [12,] -1.29196063  0.24676397  1.0488088 -1.51012476
#> [13,] -1.11864110  0.95832813  1.0488088 -1.30887353
#> [14,] -0.69200840  1.01504701  1.0488088 -0.73256319
#> [15,]  0.01460201 -0.60401927 -0.9534626  0.23252794
#> [16,]  0.01460201 -0.38230000 -0.9534626  0.15477178
#> [17,] -0.77200203 -0.69167572 -0.9534626 -1.02986615
#> [18,] -0.47869205 -0.58855048  1.0488088 -1.60343215
#> [19,]  1.53448101 -1.82605337 -0.9534626 -0.08764448
#> [20,]  2.48107231 -1.77449075 -0.9534626  0.27826685
#> [21,] -0.53202114  0.28801407  1.0488088 -0.44440802
#> attr(,"center")
#>  [1]   0.3809524   2.7142857   6.0952381 238.2809524   3.6090476   3.7142857
#>  [7] 148.9047619  18.0414286   0.4761905   3.2658095
#> attr(,"scale")
#>  [1]   0.4856209   1.6078041   1.8999940 132.3556487   0.5841470   0.7648752
#>  [7]  75.0059710   1.9393894   0.4994328   1.0931610
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 35
#> 
#> $lambda.min
#> [1] 0.5676106
#> 
#> $null.dev
#> [1] 46.81855
#> 
#> $Bias
#> [1] 5.846849
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 8.512018