Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 40.96886 36.53505 31.16399 26.56822 22.86551 19.91968 17.61363 15.84679
#>   [9] 14.53281 13.59771 12.97813 12.61991 12.47685 12.50956 12.68455 12.97342
#>  [17] 13.21765 13.11907 13.11883 13.11883 12.98068 12.84346 12.71876 12.60955
#>  [25] 12.58763 12.60053 12.84004 13.32947 13.95929 13.81644 13.46410 13.46091
#>  [33] 13.01767 11.76965 11.74444 11.72979 12.82232 17.49703 17.50215 17.68528
#>  [41] 17.53608 16.56943 16.57842 17.31409 17.31017 17.28325 18.48539 16.79901
#>  [49] 16.77543 18.33333 21.79123 29.80064 32.66241 32.55540 32.55848 32.56123
#>  [57] 32.63253 33.77743 33.77702 35.33059 35.09421 35.29742 35.29861 34.77065
#>  [65] 34.77169 34.77268 34.77364 34.80937 35.36541 35.36659 35.47373 35.55287
#>  [73] 35.55536 35.56636 35.58797 36.66095 36.66420 36.67708 36.91136 36.91395
#>  [81] 36.91644 36.91886 36.92119 37.02236 37.08045 37.18426 37.18605 37.18783
#>  [89] 37.18702 37.18868 37.19031 37.19192 37.19347 37.23266 37.22785 37.22950
#>  [97] 37.23109 37.23260 37.23403 37.23561
#> 
#> $cvse
#>   [1] 10.819382 10.066189  8.812794  7.715413  6.835657  6.111396  5.495247
#>   [8]  4.954175  4.467953  4.027222  3.631715  3.288806  3.011972  2.818083
#>  [15]  2.722365  2.731765  2.834965  2.799531  2.799432  2.799432  2.751420
#>  [22]  2.705970  2.666677  2.632278  2.598827  2.572120  2.576826  2.670462
#>  [29]  2.882335  2.783367  2.825247  2.804607  2.753173  2.779668  2.815727
#>  [36]  2.808078  2.835770  6.036573  6.034258  6.018177  5.878300  5.758216
#>  [43]  5.756664  5.741461  5.741275  5.739760  5.787212  4.511027  4.510202
#>  [50]  4.464242  5.643108  7.888376  8.268868  8.290926  8.291977  8.292928
#>  [57]  8.314457  8.846767  8.846415  9.235470  9.213356  9.298846  9.299239
#>  [64]  9.302378  9.302767  9.303156  9.303544  9.297536  9.196797  9.197123
#>  [71]  9.180660  9.215637  9.216691  9.216200  9.214845  9.749225  9.750579
#>  [78]  9.752143  9.882893  9.884166  9.885391  9.886569  9.887702  9.869638
#>  [85]  9.912169  9.898273  9.898701  9.899133  9.899725  9.900162  9.900595
#>  [92]  9.901024  9.901438  9.913699  9.913862  9.914547  9.915207  9.915842
#>  [99]  9.916453  9.917111
#> 
#> $fold
#>  [1] 10  4  8  7  6  4  5  2  9  6  2  5  1  9  3  1  3 10  7  8  1
#> 
#> $lambda
#>   [1] 5.581499187 5.205324822 4.854503351 4.527326072 4.222199447 3.937637336
#>   [7] 3.672253758 3.424756146 3.193939044 2.978678242 2.777925298 2.590702430
#>  [13] 2.416097757 2.253260854 2.101398614 1.959771381 1.827689349 1.704509204
#>  [19] 1.589630988 1.482495180 1.382579967 1.289398704 1.202497547 1.121453237
#>  [25] 1.045871042 0.975382834 0.909645296 0.848338247 0.791163089 0.737841344
#>  [31] 0.688113307 0.641736773 0.598485863 0.558149919 0.520532482 0.485450335
#>  [37] 0.452732607 0.422219945 0.393763734 0.367225376 0.342475615 0.319393904
#>  [43] 0.297867824 0.277792530 0.259070243 0.241609776 0.225326085 0.210139861
#>  [49] 0.195977138 0.182768935 0.170450920 0.158963099 0.148249518 0.138257997
#>  [55] 0.128939870 0.120249755 0.112145324 0.104587104 0.097538283 0.090964530
#>  [61] 0.084833825 0.079116309 0.073784134 0.068811331 0.064173677 0.059848586
#>  [67] 0.055814992 0.052053248 0.048545034 0.045273261 0.042221994 0.039376373
#>  [73] 0.036722538 0.034247561 0.031939390 0.029786782 0.027779253 0.025907024
#>  [79] 0.024160978 0.022532609 0.021013986 0.019597714 0.018276893 0.017045092
#>  [85] 0.015896310 0.014824952 0.013825800 0.012893987 0.012024975 0.011214532
#>  [91] 0.010458710 0.009753828 0.009096453 0.008483382 0.007911631 0.007378413
#>  [97] 0.006881133 0.006417368 0.005984859 0.005581499
#> 
#> $fit
#> $beta
#>               5.5815       5.2053       4.8545    4.5273      4.2222
#> (Intercept) 21.36667 22.408269346 23.379671492 24.285604 25.13048043
#> am           0.00000  0.000000000  0.000000000  0.000000  0.00000000
#> carb         0.00000  0.000000000  0.000000000  0.000000  0.00000000
#> cyl          0.00000  0.000000000  0.000000000  0.000000  0.00000000
#> disp         0.00000 -0.004956192 -0.009578353 -0.013889 -0.01790912
#> drat         0.00000  0.000000000  0.000000000  0.000000  0.00000000
#> gear         0.00000  0.000000000  0.000000000  0.000000  0.00000000
#> hp           0.00000  0.000000000  0.000000000  0.000000  0.00000000
#> qsec         0.00000  0.000000000  0.000000000  0.000000  0.00000000
#> vs           0.00000  0.000000000  0.000000000  0.000000  0.00000000
#> wt           0.00000  0.000000000  0.000000000  0.000000  0.00000000
#>                  3.9376      3.6723      3.4248     3.1939      2.9787
#> (Intercept) 25.91841466 26.65324476 27.33854977 27.9776675 28.57371086
#> am           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> disp        -0.02165829 -0.02515479 -0.02841563 -0.0314567 -0.03429282
#> drat         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#>                  2.7779     2.5907      2.4161      2.2533     2.1014
#> (Intercept) 29.12958289 29.6479910 30.13146013 30.58234507 31.0028419
#> am           0.00000000  0.0000000  0.00000000  0.00000000  0.0000000
#> carb         0.00000000  0.0000000  0.00000000  0.00000000  0.0000000
#> cyl          0.00000000  0.0000000  0.00000000  0.00000000  0.0000000
#> disp        -0.03693779 -0.0394045 -0.04170496 -0.04385038 -0.0458512
#> drat         0.00000000  0.0000000  0.00000000  0.00000000  0.0000000
#> gear         0.00000000  0.0000000  0.00000000  0.00000000  0.0000000
#> hp           0.00000000  0.0000000  0.00000000  0.00000000  0.0000000
#> qsec         0.00000000  0.0000000  0.00000000  0.00000000  0.0000000
#> vs           0.00000000  0.0000000  0.00000000  0.00000000  0.0000000
#> wt           0.00000000  0.0000000  0.00000000  0.00000000  0.0000000
#>                  1.9598     1.8277     1.7045     1.5896     1.4825     1.3826
#> (Intercept) 31.39499859 31.6698754 31.6698754 31.6698754 31.6698754 31.6698754
#> am           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp        -0.04771717 -0.0490251 -0.0490251 -0.0490251 -0.0490251 -0.0490251
#> drat         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> vs           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#>                 1.2894     1.2025     1.1215     1.0459     0.9754     0.9096
#> (Intercept) 31.6698754 31.6698754 31.6698754 31.6698754 31.6698754 31.6698754
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp        -0.0490251 -0.0490251 -0.0490251 -0.0490251 -0.0490251 -0.0490251
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#>                 0.8483      0.7912      0.7378      0.6881      0.6417
#> (Intercept) 31.6698754 31.49726774 31.16859663 30.49276792 29.86316000
#> am           0.0000000  0.19837713  0.66430720  1.80995940  2.87726290
#> carb         0.0000000  0.00000000 -0.08525960 -0.44192637 -0.77420405
#> cyl          0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.0490251 -0.04860833 -0.04697071 -0.04180805 -0.03699844
#> drat         0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.5985      0.5581      0.5205      0.4855      0.4527
#> (Intercept) 29.38270094 29.37756715 29.37725379 29.37707674 29.37697671
#> am           3.70847044  3.72911285  3.72963408  3.72992858  3.73009496
#> carb        -1.04395102 -1.05820332 -1.05835906 -1.05844705 -1.05849676
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.03316798 -0.03301449 -0.03301219 -0.03301089 -0.03301016
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.4222      0.3938      0.3672     0.3425      0.3194
#> (Intercept) 29.37692019 29.37688826 29.37687022 29.3768600 29.37685426
#> am           3.73018897  3.73024209  3.73027209  3.7302890  3.73029863
#> carb        -1.05852485 -1.05854072 -1.05854968 -1.0585547 -1.05855761
#> cyl          0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> disp        -0.03300974 -0.03300951 -0.03300938 -0.0330093 -0.03300926
#> drat         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#>                  0.2979      0.2778      0.2591      0.2416     0.2253
#> (Intercept) 29.37685101 24.13122584 23.33103545 23.33039371 23.3301706
#> am           3.73030404  3.08951199  2.99057919  2.99111076  2.9914824
#> carb        -1.05855923 -1.12166198 -1.13116414 -1.13136412 -1.1314878
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> disp        -0.03300924 -0.02962755 -0.02911144 -0.02910813 -0.0291062
#> drat         0.00000000  1.36738280  1.57600778  1.57606916  1.5760612
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#>                  0.2101      0.1960       0.1828      0.1705      0.1590
#> (Intercept) 23.33015184 23.33021942 17.168600653 19.36282766 18.57793240
#> am           2.99173357  2.99189826  3.105197223  3.31608986  2.61131388
#> carb        -1.13156186 -1.13160472 -0.360667115  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.000000000  0.00000000  0.00000000
#> disp        -0.02910512 -0.02910456 -0.009792666  0.00000000  0.00000000
#> drat         1.57602599  1.57598489  0.000000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.000000000  0.01335230  0.69504533
#> hp           0.00000000  0.00000000  0.000000000 -0.02067292 -0.02423014
#> qsec         0.00000000  0.00000000  0.854218733  0.76770373  0.69208219
#> vs           0.00000000  0.00000000  0.000000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000 -3.211717676 -3.50140793 -3.36192327
#>                  0.1482     0.1383      0.1289     0.1202      0.1121
#> (Intercept) 18.58130036 18.5839474 18.58602845 18.5876669 18.58896009
#> am           2.61050656  2.6098862  2.60940787  2.6090376  2.60874963
#> carb         0.00000000  0.0000000  0.00000000  0.0000000  0.00000000
#> cyl          0.00000000  0.0000000  0.00000000  0.0000000  0.00000000
#> disp         0.00000000  0.0000000  0.00000000  0.0000000  0.00000000
#> drat         0.00000000  0.0000000  0.00000000  0.0000000  0.00000000
#> gear         0.69536019  0.6956109  0.69581184  0.6959737  0.69610457
#> hp          -0.02423573 -0.0242403 -0.02424404 -0.0242471 -0.02424961
#> qsec         0.69189150  0.6917396  0.69161835  0.6915214  0.69144373
#> vs           0.00000000  0.0000000  0.00000000  0.0000000  0.00000000
#> wt          -3.36190762 -3.3618810 -3.36184873 -3.3618144 -3.36178054
#>                  0.1046      0.0975      0.0910      0.0848      0.0791
#> (Intercept) 18.58998368 18.59079648 18.59144404 18.59196163 18.59237664
#> am           2.60852470  2.60834818  2.60820900  2.60809879  2.60801114
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.69621076  0.69629714  0.69636754  0.69642502  0.69647202
#> hp          -0.02425166 -0.02425334 -0.02425472 -0.02425586 -0.02425679
#> qsec         0.69138131  0.69133103  0.69129044  0.69125759  0.69123096
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -3.36174852 -3.36171919 -3.36169293 -3.36166979 -3.36164966
#>                  0.0738      0.0688     0.0642      0.0598      0.0558
#> (Intercept) 18.59271039 18.59297953 18.5931971 18.59337341 18.59351655
#> am           2.60794117  2.60788511  2.6078400  2.60780371  2.60777434
#> carb         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> drat         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> gear         0.69651048  0.69654198  0.6965678  0.69658900  0.69660638
#> hp          -0.02425755 -0.02425818 -0.0242587 -0.02425912 -0.02425947
#> qsec         0.69120932  0.69119172  0.6911774  0.69116566  0.69115609
#> vs           0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> wt          -3.36163233 -3.36161752 -3.3616049 -3.36159433 -3.36158540
#>                  0.0521      0.0485      0.0453      0.0422      0.0394
#> (Intercept) 18.59363297 13.18594257 13.18112222 13.17706975 13.17368945
#> am           2.60775055  2.70370655  2.70406689  2.70438664  2.70466962
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.39388051  0.39405320  0.39418959  0.39429485
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.69662066  0.97103274  0.97084673  0.97066796  0.97049718
#> hp          -0.02425976 -0.02710306 -0.02709662 -0.02709077 -0.02708547
#> qsec         0.69114827  0.86449052  0.86474816  0.86496987  0.86515974
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -3.36157791 -3.60564009 -3.60604959 -3.60640215 -3.60670426
#>                  0.0367      0.0342      0.0319      0.0298     0.0278
#> (Intercept) 12.70969095 12.70669412 12.70416806 12.70206212 12.7003291
#> am           2.84312914  2.84376144  2.84431468  2.84479793  2.8452193
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> cyl          0.50865945  0.50894901  0.50918797  0.50938305  0.5095402
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> gear         0.91708381  0.91671121  0.91637365  0.91606787  0.9157910
#> hp          -0.02714676 -0.02714188 -0.02713737 -0.02713321 -0.0271294
#> qsec         0.83763175  0.83773720  0.83783023  0.83791161  0.8379822
#> vs           0.56209278  0.56300608  0.56378448  0.56444599  0.5650063
#> wt          -3.56846252 -3.56869883 -3.56890664 -3.56908827 -3.5692460
#>                  0.0259      0.0242      0.0225      0.0210      0.0196
#> (Intercept) 12.69892542 12.69781100 12.69694928 12.69630700 12.69585412
#> am           2.84558583  2.84590395  2.84617927  2.84641679  2.84662099
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.50966458  0.50976089  0.50983318  0.50988505  0.50991963
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.91554039  0.91531376  0.91510897  0.91492411  0.91475741
#> hp          -0.02712592 -0.02712274 -0.02711986 -0.02711724 -0.02711488
#> qsec         0.83804277  0.83809428  0.83813756  0.83817343  0.83820268
#> vs           0.56547890  0.56587575  0.56620712  0.56648201  0.56670826
#> wt          -3.56938187 -3.56949812 -3.56959666 -3.56967936 -3.56974796
#>                   0.0183      0.0170       0.0159       0.0148       0.0138
#> (Intercept) 12.486083505 12.48615651 12.486279989 12.486450000 12.486662636
#> am           2.948719500  2.94890242  2.949075567  2.949239236  2.949393728
#> carb         0.178890969  0.17907764  0.179254510  0.179422078  0.179580800
#> cyl          0.493091294  0.49298073  0.492870645  0.492761147  0.492652359
#> disp         0.003800618  0.00380582  0.003810773  0.003815485  0.003819965
#> drat        -0.145300122 -0.14555076 -0.145789580 -0.146016990 -0.146233387
#> gear         0.862514708  0.86240433  0.862296869  0.862192293  0.862090603
#> hp          -0.030760515 -0.03076257 -0.030764475 -0.030766240 -0.030767873
#> qsec         0.904864584  0.90497815  0.905085000  0.905185451  0.905279817
#> vs           0.660699210  0.66065668  0.660611118  0.660562493  0.660510859
#> wt          -3.909301484 -3.90979808 -3.910268970 -3.910715138 -3.911137567
#>                   0.0129       0.0120       0.0112      0.0105       0.0098
#> (Intercept) 12.486914077 12.487200618 12.487518677 12.48786482 12.488235755
#> am           2.949539346  2.949676399  2.949805197  2.94992606  2.950039298
#> carb         0.179731098  0.179873370  0.180007986  0.18013530  0.180255649
#> cyl          0.492544418  0.492437471  0.492331667  0.49222715  0.492124069
#> disp         0.003824221  0.003828262  0.003832096  0.00383573  0.003839172
#> drat        -0.146439169 -0.146634728 -0.146820453 -0.14699672 -0.147163918
#> gear         0.861991811  0.861895933  0.861802986  0.86171298  0.861625932
#> hp          -0.030769383 -0.030770778 -0.030772064 -0.03077325 -0.030774340
#> qsec         0.905368396  0.905451472  0.905529314  0.90560218  0.905670331
#> vs           0.660456336  0.660399090  0.660339324  0.66027727  0.660213165
#> wt          -3.911537215 -3.911915018 -3.912271890 -3.91260872 -3.912926391
#>                   0.0091       0.0085       0.0079       0.0074       0.0069
#> (Intercept) 12.488628363 12.489039678 12.489466902 12.489907404 12.490358716
#> am           2.950145235  2.950244189  2.950336473  2.950422401  2.950502281
#> carb         0.180369351  0.180476712  0.180578025  0.180673574  0.180763628
#> cyl          0.492022554  0.491922731  0.491824720  0.491728627  0.491634547
#> disp         0.003842431  0.003845514  0.003848428  0.003851181  0.003853779
#> drat        -0.147322400 -0.147472530 -0.147614658 -0.147749124 -0.147876261
#> gear         0.861541836  0.861460691  0.861382482  0.861307191  0.861234789
#> hp          -0.030775342 -0.030776263 -0.030777106 -0.030777879 -0.030778584
#> qsec         0.905733996  0.905793409  0.905848794  0.905900364  0.905948325
#> vs           0.660147269  0.660079836  0.660011125  0.659941387  0.659870868
#> wt          -3.913225738 -3.913507589 -3.913772746 -3.914021987 -3.914256064
#>                   0.0064       0.0060       0.0056
#> (Intercept) 12.490818533 12.491284710 12.491755255
#> am           2.950576415  2.950645098  2.950708620
#> carb         0.180848448  0.180928284  0.181003379
#> cyl          0.491542566  0.491452759  0.491365189
#> disp         0.003856229  0.003858538  0.003860712
#> drat        -0.147996390 -0.148109823 -0.148216862
#> gear         0.861165242  0.861098509  0.861034544
#> hp          -0.030779229 -0.030779815 -0.030780349
#> qsec         0.905992874  0.906034200  0.906072486
#> vs           0.659799805  0.659728423  0.659656938
#> wt          -3.914475708 -3.914681625 -3.914874497
#> 
#> $iter
#>   [1]   0   3   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1   1   1   1   1   1   7  30  35  34  21   6   1   1
#>  [37]   1   1   1   1   1   1   1 166  12   1   1   1   1 103  63  64   1   1
#>  [55]   1   1   1   1   1   1   1   1   1   1   1   1   1   1  51   1   1   1
#>  [73]  34   1   1   1   1   1   1   1   1   1 126   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.581499187 5.205324822 4.854503351 4.527326072 4.222199447 3.937637336
#>   [7] 3.672253758 3.424756146 3.193939044 2.978678242 2.777925298 2.590702430
#>  [13] 2.416097757 2.253260854 2.101398614 1.959771381 1.827689349 1.704509204
#>  [19] 1.589630988 1.482495180 1.382579967 1.289398704 1.202497547 1.121453237
#>  [25] 1.045871042 0.975382834 0.909645296 0.848338247 0.791163089 0.737841344
#>  [31] 0.688113307 0.641736773 0.598485863 0.558149919 0.520532482 0.485450335
#>  [37] 0.452732607 0.422219945 0.393763734 0.367225376 0.342475615 0.319393904
#>  [43] 0.297867824 0.277792530 0.259070243 0.241609776 0.225326085 0.210139861
#>  [49] 0.195977138 0.182768935 0.170450920 0.158963099 0.148249518 0.138257997
#>  [55] 0.128939870 0.120249755 0.112145324 0.104587104 0.097538283 0.090964530
#>  [61] 0.084833825 0.079116309 0.073784134 0.068811331 0.064173677 0.059848586
#>  [67] 0.055814992 0.052053248 0.048545034 0.045273261 0.042221994 0.039376373
#>  [73] 0.036722538 0.034247561 0.031939390 0.029786782 0.027779253 0.025907024
#>  [79] 0.024160978 0.022532609 0.021013986 0.019597714 0.018276893 0.017045092
#>  [85] 0.015896310 0.014824952 0.013825800 0.012893987 0.012024975 0.011214532
#>  [91] 0.010458710 0.009753828 0.009096453 0.008483382 0.007911631 0.007378413
#>  [97] 0.006881133 0.006417368 0.005984859 0.005581499
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 29
#> 
#> $loss
#>   [1] 834.12667 708.53701 603.46260 515.95111 443.45368 383.77103 335.00676
#>   [8] 295.52681 263.92422 238.98864 219.67976 205.10438 194.49643 187.19972
#>  [15] 182.65290 180.37651 179.91087 179.91087 179.91087 179.91087 179.91087
#>  [22] 179.91087 179.91087 179.91087 179.91087 179.91087 179.91087 179.91087
#>  [29] 176.62824 165.67333 139.31430 125.51938 121.94995 121.94254 121.94253
#>  [36] 121.94253 121.94253 121.94253 121.94253 121.94253 121.94253 121.94253
#>  [43] 121.94253 117.03849 116.95115 116.95114 116.95114 116.95114 116.95114
#>  [50] 104.55998  98.21439  96.52879  96.52878  96.52877  96.52877  96.52877
#>  [57]  96.52876  96.52876  96.52876  96.52876  96.52876  96.52876  96.52876
#>  [64]  96.52876  96.52876  96.52876  96.52876  96.52876  95.77540  95.77538
#>  [71]  95.77537  95.77536  95.54670  95.54668  95.54667  95.54666  95.54665
#>  [78]  95.54665  95.54664  95.54664  95.54664  95.54663  95.30810  95.30808
#>  [85]  95.30807  95.30805  95.30804  95.30803  95.30802  95.30801  95.30800
#>  [92]  95.30800  95.30799  95.30798  95.30798  95.30797  95.30797  95.30796
#>  [99]  95.30796  95.30796
#> 
#> $linear.predictors
#>      5.5815   5.2053   4.8545   4.5273   4.2222   3.9376   3.6723   3.4248
#> 1  21.36667 21.61528 21.84714 22.06337 22.26502 22.45309 22.62848 22.79205
#> 2  21.36667 21.87300 22.34521 22.78559 23.19630 23.57932 23.93653 24.26966
#> 3  21.36667 21.12957 20.90846 20.70224 20.50993 20.33058 20.16331 20.00732
#> 4  21.36667 20.62404 19.93146 19.28557 18.68320 18.12143 17.59752 17.10892
#> 5  21.36667 21.29313 21.22454 21.16058 21.10093 21.04530 20.99342 20.94503
#> 6  21.36667 20.62404 19.93146 19.28557 18.68320 18.12143 17.59752 17.10892
#> 7  21.36667 21.68120 21.97453 22.24809 22.50321 22.74114 22.96304 23.16998
#> 8  21.36667 21.71044 22.03104 22.33003 22.60888 22.86893 23.11145 23.33763
#> 9  21.36667 21.57761 21.77434 21.95781 22.12891 22.28848 22.43730 22.57609
#> 10 21.36667 21.04135 20.73796 20.45502 20.19115 19.94506 19.71555 19.50152
#> 11 21.36667 21.04135 20.73796 20.45502 20.19115 19.94506 19.71555 19.50152
#> 12 21.36667 20.12842 18.97363 17.89667 16.89229 15.95560 15.08204 14.26736
#> 13 21.36667 22.01822 22.62586 23.19254 23.72103 24.21391 24.67356 25.10224
#> 14 21.36667 22.03309 22.65459 23.23421 23.77476 24.27888 24.74903 25.18749
#> 15 21.36667 22.05588 22.69865 23.29810 23.85714 24.37851 24.86474 25.31820
#> 16 21.36667 21.81303 22.22931 22.61754 22.97960 23.31725 23.63215 23.92583
#> 17 21.36667 20.90159 20.46785 20.06335 19.68611 19.33429 19.00619 18.70020
#> 18 21.36667 22.01673 22.62298 23.18837 23.71566 24.20741 24.66602 25.09371
#> 19 21.36667 21.93694 22.46877 22.96476 23.42732 23.85871 24.26102 24.63622
#> 20 21.36667 20.66865 20.01767 19.41057 18.84438 18.31635 17.82391 17.36466
#> 21 21.36667 20.91646 20.49659 20.10502 19.73984 19.39927 19.08165 18.78544
#>      3.1939   2.9787   2.7779   2.5907   2.4161   2.2533    2.1014    1.9598
#> 1  22.94459 23.08686 23.21954 23.34327 23.45867 23.56629 23.666650 23.760251
#> 2  24.58034 24.87009 25.14030 25.39231 25.62732 25.84650 26.050912 26.241544
#> 3  19.86184 19.72616 19.59963 19.48163 19.37158 19.26895 19.173233 19.083968
#> 4  16.65325 16.22830 15.83198 15.46237 15.11768 14.79621 14.496410 14.216816
#> 5  20.89991 20.85783 20.81858 20.78198 20.74784 20.71601 20.686322 20.658635
#> 6  16.65325 16.22830 15.83198 15.46237 15.11768 14.79621 14.496410 14.216816
#> 7  23.36297 23.54295 23.71081 23.86735 24.01334 24.14950 24.276471 24.394889
#> 8  23.54856 23.74528 23.92874 24.09984 24.25940 24.40821 24.546993 24.676421
#> 9  22.70552 22.82623 22.93881 23.04380 23.14171 23.23302 23.318181 23.397600
#> 10 19.30191 19.11575 18.94214 18.78023 18.62923 18.48841 18.357081 18.234602
#> 11 19.30191 19.11575 18.94214 18.78023 18.62923 18.48841 18.357081 18.234602
#> 12 13.50758 12.79901 12.13820 11.52192 10.94718 10.41117  9.911291  9.445099
#> 13 25.50202 25.87487 26.22258 26.54686 26.84928 27.13132 27.394353 27.639657
#> 14 25.59640 25.97774 26.33339 26.66507 26.97439 27.26287 27.531906 27.782809
#> 15 25.74110 26.13549 26.50331 26.84633 27.16624 27.46458 27.742822 28.002308
#> 16 24.19972 24.45514 24.69335 24.91551 25.12269 25.31591 25.496113 25.664166
#> 17 18.41483 18.14869 17.90049 17.66902 17.45315 17.25183 17.064078 16.888978
#> 18 25.49259 25.86458 26.21150 26.53504 26.83677 27.11817 27.380597 27.625342
#> 19 24.98613 25.31246 25.61680 25.90062 26.16532 26.41217 26.642393 26.857095
#> 20 16.93636 16.53693 16.16442 15.81701 15.49302 15.19086 14.909071 14.646271
#> 21 18.50920 18.25157 18.01131 17.78724 17.57827 17.38338 17.201631 17.032130
#>       1.8277    1.7045    1.5896    1.4825    1.3826    1.2894    1.2025
#> 1  23.825859 23.825859 23.825859 23.825859 23.825859 23.825859 23.825859
#> 2  26.375164 26.375164 26.375164 26.375164 26.375164 26.375164 26.375164
#> 3  19.021399 19.021399 19.021399 19.021399 19.021399 19.021399 19.021399
#> 4  14.020839 14.020839 14.020839 14.020839 14.020839 14.020839 14.020839
#> 5  20.639228 20.639228 20.639228 20.639228 20.639228 20.639228 20.639228
#> 6  14.020839 14.020839 14.020839 14.020839 14.020839 14.020839 14.020839
#> 7  24.477893 24.477893 24.477893 24.477893 24.477893 24.477893 24.477893
#> 8  24.767141 24.767141 24.767141 24.767141 24.767141 24.767141 24.767141
#> 9  23.453268 23.453268 23.453268 23.453268 23.453268 23.453268 23.453268
#> 10 18.148752 18.148752 18.148752 18.148752 18.148752 18.148752 18.148752
#> 11 18.148752 18.148752 18.148752 18.148752 18.148752 18.148752 18.148752
#> 12  9.118329  9.118329  9.118329  9.118329  9.118329  9.118329  9.118329
#> 13 27.811600 27.811600 27.811600 27.811600 27.811600 27.811600 27.811600
#> 14 27.958675 27.958675 27.958675 27.958675 27.958675 27.958675 27.958675
#> 15 28.184191 28.184191 28.184191 28.184191 28.184191 28.184191 28.184191
#> 16 25.781961 25.781961 25.781961 25.781961 25.781961 25.781961 25.781961
#> 17 16.766245 16.766245 16.766245 16.766245 16.766245 16.766245 16.766245
#> 18 27.796892 27.796892 27.796892 27.796892 27.796892 27.796892 27.796892
#> 19 27.007588 27.007588 27.007588 27.007588 27.007588 27.007588 27.007588
#> 20 14.462065 14.462065 14.462065 14.462065 14.462065 14.462065 14.462065
#> 21 16.913320 16.913320 16.913320 16.913320 16.913320 16.913320 16.913320
#>       1.1215    1.0459    0.9754    0.9096    0.8483    0.7912    0.7378
#> 1  23.825859 23.825859 23.825859 23.825859 23.825859 23.918312 23.976552
#> 2  26.375164 26.375164 26.375164 26.375164 26.375164 26.445945 26.674807
#> 3  19.021399 19.021399 19.021399 19.021399 19.021399 18.956318 18.964894
#> 4  14.020839 14.020839 14.020839 14.020839 14.020839 13.998268 14.088622
#> 5  20.639228 20.639228 20.639228 20.639228 20.639228 20.560393 20.514927
#> 6  14.020839 14.020839 14.020839 14.020839 14.020839 13.998268 13.918102
#> 7  24.477893 24.477893 24.477893 24.477893 24.477893 24.366425 24.107474
#> 8  24.767141 24.767141 24.767141 24.767141 24.767141 24.653214 24.384601
#> 9  23.453268 23.453268 23.453268 23.453268 23.453268 23.350511 22.955267
#> 10 18.148752 18.148752 18.148752 18.148752 18.148752 18.091090 17.958296
#> 11 18.148752 18.148752 18.148752 18.148752 18.148752 18.091090 17.958296
#> 12  9.118329  9.118329  9.118329  9.118329  9.118329  9.137435  9.221031
#> 13 27.811600 27.811600 27.811600 27.811600 27.811600 27.870169 28.051049
#> 14 27.958675 27.958675 27.958675 27.958675 27.958675 28.015994 28.106702
#> 15 28.184191 28.184191 28.184191 28.184191 28.184191 28.239592 28.408027
#> 16 25.781961 25.781961 25.781961 25.781961 25.781961 25.659407 25.442155
#> 17 16.766245 16.766245 16.766245 16.766245 16.766245 16.720335 16.718981
#> 18 27.796892 27.796892 27.796892 27.796892 27.796892 27.855587 28.036958
#> 19 27.007588 27.007588 27.007588 27.007588 27.007588 27.072992 27.195470
#> 20 14.462065 14.462065 14.462065 14.462065 14.462065 14.634120 15.005146
#> 21 16.913320 16.913320 16.913320 16.913320 16.913320 17.064537 17.012643
#>       0.6881    0.6417    0.5985    0.5581    0.5205    0.4855    0.4527
#> 1  23.845733 23.723856 23.608491 23.591548 23.591501 23.591474 23.591459
#> 2  27.345531 27.970387 28.465079 28.482912 28.483212 28.483382 28.483478
#> 3  19.264364 19.543359 19.781412 19.801625 19.801749 19.801819 19.801859
#> 4  14.558016 14.995314 15.354327 15.375944 15.376147 15.376261 15.376326
#> 5  20.644030 20.764307 20.875955 20.891104 20.891152 20.891179 20.891194
#> 6  13.674163 13.446906 13.266425 13.259537 13.259429 13.259367 13.259332
#> 7  23.475674 22.887081 22.429057 22.417935 22.417647 22.417485 22.417393
#> 8  23.722341 23.105372 22.624748 22.612720 22.612419 22.612249 22.612153
#> 9  21.718033 20.565405 19.647944 19.611525 19.610974 19.610663 19.610487
#> 10 17.636328 17.336378 17.103120 17.097561 17.097414 17.097331 17.097284
#> 11 17.636328 17.336378 17.103120 17.097561 17.097414 17.097331 17.097284
#> 12  9.493358  9.747062  9.949627  9.958089  9.958209  9.958278  9.958316
#> 13 28.570507 29.054442 29.436901 29.450236 29.450469 29.450601 29.450675
#> 14 28.254005 28.391233 28.492453 28.491076 28.491147 28.491187 28.491209
#> 15 28.888248 29.335630 29.688977 29.701146 29.701362 29.701484 29.701553
#> 16 25.029694 24.645443 24.355276 24.354324 24.354131 24.354021 24.353960
#> 17 16.899267 17.067226 17.211734 17.224756 17.224829 17.224871 17.224895
#> 18 28.557965 29.043342 29.426950 29.440332 29.440566 29.440698 29.440772
#> 19 27.442929 27.673463 27.848995 27.850595 27.850710 27.850775 27.850812
#> 20 15.860395 16.657155 17.273407 17.285781 17.286172 17.286394 17.286519
#> 21 16.183092 15.410260 14.756002 14.703692 14.703346 14.703150 14.703040
#>       0.4222    0.3938    0.3672    0.3425    0.3194    0.2979   0.2778
#> 1  23.591451 23.591446 23.591443 23.591442 23.591441 23.591440 23.32647
#> 2  28.483532 28.483562 28.483580 28.483590 28.483595 28.483598 28.16372
#> 3  19.801881 19.801894 19.801901 19.801905 19.801907 19.801909 19.57719
#> 4  15.376362 15.376383 15.376395 15.376401 15.376405 15.376407 15.52924
#> 5  20.891203 20.891208 20.891210 20.891212 20.891213 20.891213 20.11734
#> 6  13.259313 13.259302 13.259295 13.259292 13.259290 13.259289 13.36796
#> 7  22.417341 22.417312 22.417295 22.417286 22.417280 22.417277 22.58718
#> 8  22.612098 22.612068 22.612050 22.612041 22.612035 22.612032 23.07648
#> 9  19.610388 19.610331 19.610300 19.610282 19.610272 19.610266 20.03914
#> 10 17.097258 17.097243 17.097235 17.097230 17.097227 17.097226 16.79283
#> 11 17.097258 17.097243 17.097235 17.097230 17.097227 17.097226 16.79283
#> 12  9.958338  9.958351  9.958357  9.958361  9.958364  9.958365 10.11805
#> 13 29.450717 29.450741 29.450755 29.450762 29.450766 29.450769 29.34631
#> 14 28.491222 28.491229 28.491233 28.491235 28.491237 28.491237 29.47581
#> 15 29.701591 29.701613 29.701626 29.701633 29.701637 29.701639 29.76291
#> 16 24.353925 24.353905 24.353894 24.353888 24.353884 24.353882 24.51061
#> 17 17.224908 17.224916 17.224920 17.224922 17.224924 17.224924 17.18838
#> 18 29.440814 29.440838 29.440852 29.440859 29.440864 29.440866 29.33742
#> 19 27.850833 27.850844 27.850851 27.850855 27.850857 27.850858 27.31487
#> 20 17.286589 17.286629 17.286652 17.286665 17.286672 17.286676 18.10517
#> 21 14.702977 14.702942 14.702922 14.702911 14.702904 14.702901 14.17008
#>      0.2591   0.2416   0.2253   0.2101   0.1960    0.1828   0.1705    0.1590
#> 1  23.28556 23.28542 23.28535 23.28532 23.28531 22.910043 23.92102 23.887546
#> 2  28.11405 28.11433 28.11453 28.11467 28.11476 27.301348 26.97345 26.796012
#> 3  19.54322 19.54343 19.54355 19.54363 19.54367 20.561766 20.79600 20.643247
#> 4  15.55301 15.55336 15.55356 15.55367 15.55372 16.412401 16.80660 16.637016
#> 5  19.99958 19.99965 19.99972 19.99977 19.99981 20.764343 20.64033 20.480551
#> 6  13.38525 13.38519 13.38514 13.38510 13.38507 14.265565 13.99842 13.687199
#> 7  22.61353 22.61320 22.61298 22.61284 22.61276 21.849678 22.31910 22.972954
#> 8  23.14777 23.14743 23.14720 23.14705 23.14695 24.513157 24.00329 24.314874
#> 9  20.10525 20.10461 20.10418 20.10391 20.10374 18.668575 18.87760 19.477894
#> 10 16.74695 16.74681 16.74673 16.74667 16.74664 15.177497 14.78907 14.660845
#> 11 16.74695 16.74681 16.74673 16.74667 16.74664 16.440325 16.13309 15.942316
#> 12 10.14314 10.14341 10.14355 10.14363 10.14366  9.023127  9.64705  9.551421
#> 13 29.32949 29.32969 29.32984 29.32994 29.33000 28.708308 28.61201 28.448847
#> 14 29.62527 29.62531 29.62532 29.62531 29.62530 29.444366 30.22043 30.097316
#> 15 29.77138 29.77156 29.77169 29.77178 29.77184 30.322323 30.24081 29.997775
#> 16 24.53482 24.53460 24.53445 24.53436 24.53430 24.807867 24.12839 23.874168
#> 17 17.18326 17.18341 17.18350 17.18355 17.18358 17.216030 17.55588 17.453363
#> 18 29.32076 29.32096 29.32111 29.32120 29.32127 29.069571 29.10229 28.945270
#> 19 27.23234 27.23237 27.23243 27.23248 27.23253 28.198149 28.08620 28.536066
#> 20 18.22960 18.23011 18.23040 18.23057 18.23066 17.598930 17.32027 17.645610
#> 21 14.08883 14.08833 14.08804 14.08788 14.08779 15.446630 14.52870 14.649709
#>       0.1482    0.1383    0.1289    0.1202    0.1121   0.1046    0.0975
#> 1  23.887653 23.887749 23.887834 23.887907 23.887970 23.88802 23.888070
#> 2  26.795800 26.795639 26.795516 26.795423 26.795352 26.79530 26.795254
#> 3  20.643288 20.643317 20.643336 20.643350 20.643359 20.64336 20.643368
#> 4  16.637159 16.637264 16.637341 16.637398 16.637440 16.63747 16.637493
#> 5  20.480475 20.480414 20.480366 20.480328 20.480297 20.48027 20.480252
#> 6  13.687177 13.687145 13.687109 13.687070 13.687033 13.68700 13.686965
#> 7  22.973470 22.973883 22.974215 22.974482 22.974698 22.97487 22.975014
#> 8  24.314653 24.314473 24.314329 24.314212 24.314119 24.31404 24.313983
#> 9  19.478398 19.478797 19.479115 19.479369 19.479572 19.47974 19.479868
#> 10 14.660897 14.660938 14.660972 14.660998 14.661019 14.66104 14.661049
#> 11 15.942324 15.942326 15.942324 15.942319 15.942313 15.94231 15.942300
#> 12  9.551218  9.551071  9.550967  9.550892  9.550838  9.55080  9.550773
#> 13 28.448620 28.448449 28.448319 28.448221 28.448146 28.44809 28.448046
#> 14 30.097339 30.097361 30.097380 30.097396 30.097411 30.09742 30.097434
#> 15 29.997465 29.997223 29.997034 29.996884 29.996767 29.99667 29.996599
#> 16 23.874161 23.874143 23.874118 23.874090 23.874062 23.87404 23.874010
#> 17 17.453591 17.453768 17.453905 17.454011 17.454093 17.45416 17.454208
#> 18 28.945147 28.945055 28.944986 28.944935 28.944895 28.94487 28.944843
#> 19 28.536370 28.536607 28.536792 28.536937 28.537052 28.53714 28.537216
#> 20 17.645554 17.645509 17.645474 17.645447 17.645427 17.64541 17.645398
#> 21 14.649243 14.648869 14.648570 14.648330 14.648137 14.64798 14.647856
#>       0.0910    0.0848    0.0791    0.0738   0.0688    0.0642    0.0598
#> 1  23.888109 23.888141 23.888169 23.888191 23.88821 23.888226 23.888239
#> 2  26.795221 26.795196 26.795175 26.795160 26.79515 26.795137 26.795129
#> 3  20.643370 20.643372 20.643372 20.643372 20.64337 20.643372 20.643372
#> 4  16.637510 16.637522 16.637531 16.637538 16.63754 16.637546 16.637549
#> 5  20.480236 20.480223 20.480212 20.480203 20.48020 20.480191 20.480186
#> 6  13.686936 13.686911 13.686889 13.686870 13.68685 13.686840 13.686828
#> 7  22.975130 22.975224 22.975301 22.975363 22.97541 22.975457 22.975491
#> 8  24.313934 24.313894 24.313862 24.313836 24.31382 24.313798 24.313784
#> 9  19.479975 19.480061 19.480131 19.480189 19.48024 19.480273 19.480304
#> 10 14.661060 14.661069 14.661075 14.661081 14.66109 14.661089 14.661092
#> 11 15.942294 15.942288 15.942282 15.942278 15.94227 15.942270 15.942267
#> 12  9.550754  9.550741  9.550731  9.550724  9.55072  9.550716  9.550714
#> 13 28.448012 28.447986 28.447965 28.447949 28.44794 28.447927 28.447919
#> 14 30.097442 30.097450 30.097456 30.097461 30.09747 30.097469 30.097472
#> 15 29.996539 29.996492 29.996454 29.996423 29.99640 29.996378 29.996361
#> 16 23.873987 23.873967 23.873949 23.873934 23.87392 23.873910 23.873900
#> 17 17.454248 17.454279 17.454303 17.454323 17.45434 17.454351 17.454362
#> 18 28.944825 28.944811 28.944801 28.944793 28.94479 28.944781 28.944777
#> 19 28.537274 28.537321 28.537358 28.537388 28.53741 28.537433 28.537449
#> 20 17.645389 17.645381 17.645375 17.645371 17.64537 17.645364 17.645362
#> 21 14.647755 14.647673 14.647607 14.647552 14.64751 14.647473 14.647443
#>       0.0558    0.0521    0.0485    0.0453    0.0422    0.0394    0.0367
#> 1  23.888250 23.888259 23.938464 23.938172 23.937912 23.937680 23.725015
#> 2  26.795122 26.795117 26.551801 26.551732 26.551681 26.551648 26.602731
#> 3  20.643371 20.643371 20.694550 20.694609 20.694658 20.694700 20.899803
#> 4  16.637551 16.637553 16.817276 16.817383 16.817470 16.817539 16.760517
#> 5  20.480182 20.480179 20.620986 20.621113 20.621220 20.621309 20.814616
#> 6  13.686818 13.686810 13.431230 13.431431 13.431619 13.431796 13.407938
#> 7  22.975519 22.975542 22.753025 22.752397 22.751847 22.751367 22.660897
#> 8  24.313773 24.313764 24.509872 24.510220 24.510520 24.510778 24.336925
#> 9  19.480330 19.480351 19.516455 19.516025 19.515640 19.515295 19.706174
#> 10 14.661094 14.661096 14.738714 14.738693 14.738671 14.738649 14.694952
#> 11 15.942265 15.942262 16.137530 16.137700 16.137842 16.137960 16.075755
#> 12  9.550712  9.550711  9.271157  9.270915  9.270713  9.270547  9.264922
#> 13 28.447913 28.447908 28.459723 28.459750 28.459774 28.459797 28.484272
#> 14 30.097474 30.097476 30.127199 30.127131 30.127069 30.127014 30.156127
#> 15 29.996348 29.996337 30.174615 30.174896 30.175139 30.175348 30.174090
#> 16 23.873892 23.873886 23.456119 23.456202 23.456293 23.456390 23.389189
#> 17 17.454370 17.454376 17.754938 17.754958 17.754962 17.754954 17.691565
#> 18 28.944774 28.944771 28.922458 28.922446 28.922438 28.922433 28.952465
#> 19 28.537462 28.537473 28.412246 28.412009 28.411802 28.411623 28.424279
#> 20 17.645360 17.645359 17.845883 17.846013 17.846119 17.846204 17.874405
#> 21 14.647419 14.647400 14.565759 14.566208 14.566610 14.566970 14.603363
#>       0.0342    0.0319    0.0298    0.0278    0.0259    0.0242    0.0225
#> 1  23.724551 23.724145 23.723790 23.723481 23.723212 23.722980 23.722780
#> 2  26.602816 26.602896 26.602971 26.603042 26.603109 26.603173 26.603233
#> 3  20.900166 20.900476 20.900739 20.900962 20.901150 20.901308 20.901439
#> 4  16.760555 16.760585 16.760610 16.760628 16.760641 16.760650 16.760655
#> 5  20.814979 20.815288 20.815550 20.815770 20.815955 20.816108 20.816234
#> 6  13.408163 13.408372 13.408567 13.408749 13.408917 13.409072 13.409216
#> 7  22.660140 22.659475 22.658893 22.658386 22.657944 22.657562 22.657232
#> 8  24.336643 24.336406 24.336204 24.336033 24.335888 24.335765 24.335659
#> 9  19.706055 19.705933 19.705811 19.705691 19.705574 19.705461 19.705354
#> 10 14.694905 14.694863 14.694824 14.694789 14.694757 14.694729 14.694703
#> 11 16.075810 16.075857 16.075897 16.075929 16.075956 16.075977 16.075993
#> 12  9.264771  9.264644  9.264539  9.264454  9.264385  9.264332  9.264291
#> 13 28.484345 28.484408 28.484462 28.484510 28.484552 28.484588 28.484619
#> 14 30.156169 30.156202 30.156228 30.156248 30.156262 30.156273 30.156281
#> 15 30.174289 30.174463 30.174615 30.174747 30.174861 30.174958 30.175041
#> 16 23.389147 23.389129 23.389131 23.389149 23.389179 23.389219 23.389267
#> 17 17.691512 17.691457 17.691401 17.691344 17.691288 17.691233 17.691178
#> 18 28.952540 28.952605 28.952661 28.952711 28.952754 28.952791 28.952824
#> 19 28.424099 28.423940 28.423800 28.423677 28.423569 28.423476 28.423394
#> 20 17.874562 17.874694 17.874804 17.874894 17.874966 17.875025 17.875070
#> 21 14.603783 14.604162 14.604502 14.604806 14.605078 14.605320 14.605535
#>       0.0210    0.0196   0.0183    0.0170    0.0159    0.0148    0.0138
#> 1  23.722609 23.722462 23.86845 23.868541 23.868631 23.868717 23.868800
#> 2  26.603290 26.603344 26.45710 26.456985 26.456877 26.456776 26.456684
#> 3  20.901547 20.901636 21.04331 21.043485 21.043653 21.043811 21.043959
#> 4  16.760657 16.760656 16.85637 16.856554 16.856724 16.856886 16.857039
#> 5  20.816337 20.816420 20.86620 20.866265 20.866323 20.866376 20.866423
#> 6  13.409347 13.409468 13.47625 13.476448 13.476637 13.476820 13.476997
#> 7  22.656948 22.656706 22.66785 22.667768 22.667692 22.667620 22.667554
#> 8  24.335569 24.335492 24.37739 24.377501 24.377606 24.377705 24.377800
#> 9  19.705252 19.705156 19.66584 19.665521 19.665211 19.664912 19.664624
#> 10 14.694679 14.694658 14.45406 14.453732 14.453415 14.453114 14.452827
#> 11 16.076005 16.076013 15.96420 15.964059 15.963924 15.963794 15.963670
#> 12  9.264262  9.264243  9.35343  9.353564  9.353694  9.353821  9.353944
#> 13 28.484647 28.484670 28.39016 28.390044 28.389935 28.389833 28.389735
#> 14 30.156286 30.156289 30.29211 30.292164 30.292216 30.292263 30.292307
#> 15 30.175111 30.175170 30.18768 30.187722 30.187762 30.187799 30.187834
#> 16 23.389320 23.389377 23.29057 23.290559 23.290555 23.290556 23.290561
#> 17 17.691125 17.691074 17.68546 17.685435 17.685408 17.685380 17.685351
#> 18 28.952853 28.952878 28.91149 28.911444 28.911402 28.911361 28.911323
#> 19 28.423324 28.423264 28.45338 28.453458 28.453534 28.453608 28.453678
#> 20 17.875104 17.875129 17.73580 17.735660 17.735531 17.735406 17.735287
#> 21 14.605726 14.605894 14.70290 14.703091 14.703270 14.703442 14.703605
#>       0.0129    0.0120    0.0112   0.0105    0.0098    0.0091    0.0085
#> 1  23.868878 23.868954 23.869026 23.86910 23.869163 23.869228 23.869290
#> 2  26.456598 26.456518 26.456445 26.45638 26.456315 26.456258 26.456206
#> 3  21.044097 21.044227 21.044347 21.04446 21.044564 21.044661 21.044750
#> 4  16.857183 16.857320 16.857449 16.85757 16.857686 16.857794 16.857895
#> 5  20.866466 20.866504 20.866537 20.86657 20.866591 20.866612 20.866630
#> 6  13.477168 13.477333 13.477492 13.47764 13.477792 13.477934 13.478070
#> 7  22.667493 22.667436 22.667385 22.66734 22.667294 22.667256 22.667221
#> 8  24.377889 24.377974 24.378054 24.37813 24.378203 24.378272 24.378338
#> 9  19.664346 19.664079 19.663822 19.66358 19.663339 19.663112 19.662894
#> 10 14.452555 14.452296 14.452050 14.45182 14.451596 14.451386 14.451188
#> 11 15.963551 15.963437 15.963328 15.96322 15.963125 15.963030 15.962939
#> 12  9.354063  9.354179  9.354291  9.35440  9.354506  9.354608  9.354706
#> 13 28.389642 28.389553 28.389470 28.38939 28.389315 28.389243 28.389176
#> 14 30.292348 30.292385 30.292419 30.29245 30.292480 30.292506 30.292530
#> 15 30.187865 30.187894 30.187920 30.18794 30.187965 30.187984 30.188001
#> 16 23.290571 23.290585 23.290602 23.29062 23.290646 23.290671 23.290700
#> 17 17.685322 17.685293 17.685263 17.68523 17.685203 17.685173 17.685143
#> 18 28.911286 28.911252 28.911220 28.91119 28.911160 28.911133 28.911107
#> 19 28.453746 28.453812 28.453875 28.45394 28.453993 28.454049 28.454102
#> 20 17.735172 17.735062 17.734956 17.73485 17.734755 17.734660 17.734569
#> 21 14.703760 14.703908 14.704048 14.70418 14.704309 14.704429 14.704544
#>       0.0079    0.0074    0.0069    0.0064    0.0060    0.0056
#> 1  23.869350 23.869408 23.869463 23.869517 23.869569 23.869618
#> 2  26.456158 26.456113 26.456073 26.456037 26.456003 26.455973
#> 3  21.044833 21.044909 21.044979 21.045043 21.045102 21.045155
#> 4  16.857991 16.858080 16.858164 16.858243 16.858316 16.858385
#> 5  20.866644 20.866656 20.866664 20.866670 20.866674 20.866675
#> 6  13.478200 13.478326 13.478446 13.478560 13.478670 13.478775
#> 7  22.667189 22.667162 22.667138 22.667116 22.667098 22.667083
#> 8  24.378400 24.378460 24.378516 24.378569 24.378620 24.378668
#> 9  19.662686 19.662487 19.662296 19.662115 19.661941 19.661776
#> 10 14.451001 14.450823 14.450656 14.450498 14.450349 14.450208
#> 11 15.962853 15.962771 15.962693 15.962618 15.962547 15.962480
#> 12  9.354802  9.354893  9.354982  9.355067  9.355149  9.355228
#> 13 28.389112 28.389051 28.388994 28.388940 28.388889 28.388841
#> 14 30.292552 30.292572 30.292591 30.292607 30.292622 30.292635
#> 15 30.188017 30.188030 30.188042 30.188053 30.188062 30.188070
#> 16 23.290730 23.290761 23.290794 23.290829 23.290864 23.290900
#> 17 17.685113 17.685083 17.685054 17.685025 17.684996 17.684968
#> 18 28.911082 28.911059 28.911037 28.911017 28.910998 28.910979
#> 19 28.454154 28.454202 28.454249 28.454294 28.454337 28.454378
#> 20 17.734482 17.734398 17.734317 17.734239 17.734165 17.734093
#> 21 14.704652 14.704754 14.704851 14.704943 14.705030 14.705111
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 22.8 21.4 18.7 18.1 14.3 24.4 22.8 19.2 16.4 17.3 10.4 32.4 30.4 33.9
#> [16] 21.5 15.2 27.3 30.4 15.8 15.0
#> 
#> $X
#>               am       carb         cyl       disp       drat       gear
#>  [1,]  1.1547005  0.8789244  0.05299989 -0.4405971  0.5048895  0.4082483
#>  [2,]  1.1547005 -0.9072768 -1.05999788 -0.8973392  0.4100534  0.4082483
#>  [3,] -0.8660254 -0.9072768  0.05299989  0.4201859 -1.0504230 -1.0206207
#>  [4,] -0.8660254 -0.3118764  1.16599767  1.3161030 -0.9176525 -1.0206207
#>  [5,] -0.8660254 -0.9072768  0.05299989  0.1303304 -1.6573743 -1.0206207
#>  [6,] -0.8660254  0.8789244  1.16599767  1.3161030 -0.8038491 -1.0206207
#>  [7,] -0.8660254 -0.3118764 -1.05999788 -0.5574177  0.1065777  0.4082483
#>  [8,] -0.8660254 -0.3118764 -1.05999788 -0.6092403  0.5428239  0.4082483
#>  [9,] -0.8660254  0.8789244  0.05299989 -0.3738425  0.5428239  0.4082483
#> [10,] -0.8660254  0.2835240  1.16599767  0.5765323 -1.0693903 -1.0206207
#> [11,] -0.8660254  0.2835240  1.16599767  0.5765323 -1.0693903 -1.0206207
#> [12,] -0.8660254  0.8789244  1.16599767  2.1944531 -1.2021608 -1.0206207
#> [13,]  1.1547005 -0.9072768 -1.05999788 -1.1546957  0.8462996  0.4082483
#> [14,]  1.1547005 -0.3118764 -1.05999788 -1.1810462  2.4585138  0.4082483
#> [15,]  1.1547005 -0.9072768 -1.05999788 -1.2214503  1.1118407  0.4082483
#> [16,] -0.8660254 -0.9072768 -1.05999788 -0.7910588  0.1255450 -1.0206207
#> [17,] -0.8660254 -0.3118764  1.16599767  0.8242270 -0.9176525 -1.0206207
#> [18,]  1.1547005 -0.9072768 -1.05999788 -1.1520607  0.8462996  0.4082483
#> [19,]  1.1547005 -0.3118764 -1.05999788 -1.0106463  0.2583156  1.8371173
#> [20,]  1.1547005  0.8789244  1.16599767  1.2370515  1.1118407  1.8371173
#> [21,]  1.1547005  3.2605259  1.16599767  0.7978765 -0.1779306  1.8371173
#>               hp        qsec         vs         wt
#>  [1,] -0.3802451 -0.88079663 -1.1547005 -0.4382704
#>  [2,] -0.6099363  0.22233701  0.8660254 -0.7690603
#>  [3,] -0.3802451  0.64819791  0.8660254  0.2177963
#>  [4,]  0.4979860 -0.59346880 -1.1547005  0.4658887
#>  [5,] -0.4478014  1.04840453  0.8660254  0.4879414
#>  [6,]  1.4437733 -1.19890959 -1.1547005  0.6092310
#>  [7,] -1.0287850  0.93552574  0.8660254  0.1902305
#>  [8,] -0.5829138  2.42347345  0.8660254  0.1461251
#>  [9,] -0.2045989  0.06328053  0.8660254  0.4658887
#> [10,]  0.5655422 -0.39849634 -1.1547005  1.1605476
#> [11,]  0.5655422 -0.29587926 -1.1547005  0.7856523
#> [12,]  1.0384359 -0.18300046 -1.1547005  2.6535128
#> [13,] -0.9747400  0.66359047  0.8660254 -0.9013763
#> [14,] -1.1638975  0.17615933  0.8660254 -1.5464167
#> [15,] -0.9882513  0.88421720  0.8660254 -1.3038374
#> [16,] -0.5558913  0.94065659  0.8660254 -0.6091785
#> [17,]  0.1602048 -0.44980488 -1.1547005  0.4603756
#> [18,] -0.9747400  0.37113178  0.8660254 -1.1935741
#> [19,] -0.3397114 -0.65503905  0.8660254 -1.6588852
#> [20,]  1.7004870 -1.88644404 -1.1547005  0.1681778
#> [21,]  2.6597856 -1.83513550 -1.1547005  0.6092310
#> attr(,"center")
#>  [1]   0.4285714   2.5238095   5.9047619 210.1619048   3.6338095   3.7142857
#>  [7] 138.1428571  18.1766667   0.5714286   3.0174762
#> attr(,"scale")
#>  [1]   0.4948717   1.6795421   1.7969488 113.8498237   0.5272252   0.6998542
#>  [7]  74.0124093   1.9489932   0.4948717   0.9069200
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 36
#> 
#> $lambda.min
#> [1] 0.4854503
#> 
#> $null.dev
#> [1] 39.72032
#> 
#> $Bias
#> [1] 4.689707
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 8.781669