Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 31.568998 29.525083 26.601217 22.985446 20.098921 17.829250 16.079851
#>   [8] 14.767819 13.822066 13.181715 12.794708 12.616591 12.609477 12.745111
#>  [15] 12.925244 12.948646 12.658297 12.269158 11.802111 11.550085 11.247057
#>  [22] 10.871882 10.568439 10.326108 10.147860 10.071984 10.000148  9.949984
#>  [29]  9.917855  9.885005  8.312140  8.263510  8.229443  8.210401  8.244767
#>  [36]  8.291919  8.415131  8.534886  8.664219  8.800685  9.657327  9.808147
#>  [43]  9.807383 10.163674 11.605255 11.808118 11.700567 11.572587 11.988642
#>  [50] 12.138850 18.097816 20.050693 20.219634 21.020749 23.022234 24.294611
#>  [57] 24.308735 24.686669 24.496597 24.969181 31.951269 33.799200 33.601814
#>  [64] 33.575916 33.394657 33.396106 33.240589 33.241777 33.177760 33.178219
#>  [71] 33.178740 33.191869 33.193830 34.327915 34.328408 34.328891 35.032342
#>  [78] 35.033374 35.034378 35.035353 34.809387 34.804911 34.805869 35.330882
#>  [85] 35.331274 35.360046 35.361735 35.363354 35.364905 35.366392 35.367816
#>  [92] 35.369180 35.370421 35.372240 35.373419 35.374546 35.375625 35.376674
#>  [99] 35.376075 35.376998
#> 
#> $cvse
#>   [1]  9.934220  9.743061  9.198857  8.093009  7.199001  6.469974  5.868919
#>   [8]  5.368297  4.949112  4.599566  4.313465  4.088486  3.924390  3.820629
#>  [15]  3.779975  3.789120  3.732628  3.655799  3.545091  3.490628  3.425533
#>  [22]  3.352640  3.290763  3.238431  3.194693  3.117558  3.033999  2.959527
#>  [29]  2.893817  2.836815  2.232669  2.214254  2.195719  2.185297  2.184882
#>  [36]  2.184362  2.202179  2.223481  2.247428  2.273977  2.650010  2.651749
#>  [43]  2.651580  2.640335  3.420635  3.397002  3.396073  3.355161  3.328461
#>  [50]  3.330326  5.955540  6.050395  6.063144  6.000066  5.972697  6.868201
#>  [57]  6.871077  7.002867  7.020859  7.615794  8.955379  9.917834  9.944507
#>  [64]  9.953143  9.976175  9.976597  9.996521  9.996962 10.000856 10.001230
#>  [71] 10.001596 10.000845 10.000683 10.115417 10.115794 10.116165 10.166793
#>  [78] 10.167199 10.167597 10.167989 10.200234 10.201318 10.201680 10.468091
#>  [85] 10.469136 10.466295 10.467063 10.467799 10.468504 10.469178 10.469823
#>  [92] 10.470441 10.471040 10.471528 10.472069 10.472587 10.473081 10.473552
#>  [99] 10.474050 10.474481
#> 
#> $fold
#>  [1]  5  1  2  8  7  3  1  7 10  9  2  6 10  9  6  4  5  3  4  1  8
#> 
#> $lambda
#>   [1] 4.650114832 4.336712656 4.044432738 3.771851507 3.517641340 3.280564086
#>   [7] 3.059465047 2.853267342 2.660966673 2.481626425 2.314373110 2.158392108
#>  [13] 2.012923704 1.877259383 1.750738384 1.632744476 1.522702963 1.420077880
#>  [19] 1.324369384 1.235111320 1.151868950 1.074236838 1.001836871 0.934316419
#>  [25] 0.871346619 0.812620773 0.757852853 0.706776107 0.659141763 0.614717814
#>  [31] 0.573287891 0.534650206 0.498616571 0.465011483 0.433671266 0.404443274
#>  [37] 0.377185151 0.351764134 0.328056409 0.305946505 0.285326734 0.266096667
#>  [43] 0.248162643 0.231437311 0.215839211 0.201292370 0.187725938 0.175073838
#>  [49] 0.163274448 0.152270296 0.142007788 0.132436938 0.123511132 0.115186895
#>  [55] 0.107423684 0.100183687 0.093431642 0.087134662 0.081262077 0.075785285
#>  [61] 0.070677611 0.065914176 0.061471781 0.057328789 0.053465021 0.049861657
#>  [67] 0.046501148 0.043367127 0.040444327 0.037718515 0.035176413 0.032805641
#>  [73] 0.030594650 0.028532673 0.026609667 0.024816264 0.023143731 0.021583921
#>  [79] 0.020129237 0.018772594 0.017507384 0.016327445 0.015227030 0.014200779
#>  [85] 0.013243694 0.012351113 0.011518690 0.010742368 0.010018369 0.009343164
#>  [91] 0.008713466 0.008126208 0.007578529 0.007067761 0.006591418 0.006147178
#>  [97] 0.005732879 0.005346502 0.004986166 0.004650115
#> 
#> $fit
#> $beta
#>               4.6501     4.3367    4.0444    3.7719   3.5176   3.2806    3.0595
#> (Intercept) 19.84762 21.7276480 23.480969 25.116123 26.64107 28.06325 29.389568
#> am           0.00000  0.0000000  0.000000  0.000000  0.00000  0.00000  0.000000
#> carb         0.00000  0.0000000  0.000000  0.000000  0.00000  0.00000  0.000000
#> cyl          0.00000  0.0000000  0.000000  0.000000  0.00000  0.00000  0.000000
#> disp         0.00000  0.0000000  0.000000  0.000000  0.00000  0.00000  0.000000
#> drat         0.00000  0.0000000  0.000000  0.000000  0.00000  0.00000  0.000000
#> gear         0.00000  0.0000000  0.000000  0.000000  0.00000  0.00000  0.000000
#> hp           0.00000  0.0000000  0.000000  0.000000  0.00000  0.00000  0.000000
#> qsec         0.00000  0.0000000  0.000000  0.000000  0.00000  0.00000  0.000000
#> vs           0.00000  0.0000000  0.000000  0.000000  0.00000  0.00000  0.000000
#> wt           0.00000 -0.5842314 -1.129088 -1.637222 -2.11111 -2.55306 -2.965224
#>                2.8533    2.6610    2.4816    2.3144    2.1584    2.0129
#> (Intercept) 30.626502 31.780070 32.855892 33.859207 34.794902 35.667534
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -3.349609 -3.708088 -4.042407 -4.354194 -4.644967 -4.916143
#>                1.8773    1.7507      1.6327    1.5227     1.4201     1.3244
#> (Intercept) 36.481353 37.240324 36.77497012 35.914719 34.3180284 32.8288634
#> am           0.000000  0.000000  0.00000000  0.000000  0.0000000  0.0000000
#> carb         0.000000  0.000000  0.00000000  0.000000  0.0000000  0.0000000
#> cyl          0.000000  0.000000  0.00000000  0.000000  0.0000000  0.0000000
#> disp         0.000000  0.000000  0.00000000  0.000000  0.0000000  0.0000000
#> drat         0.000000  0.000000  0.00000000  0.000000  0.0000000  0.0000000
#> gear         0.000000  0.000000  0.00000000  0.000000  0.0000000  0.0000000
#> hp           0.000000  0.000000  0.00000000  0.000000  0.0000000  0.0000000
#> qsec         0.000000  0.000000  0.05961406  0.132525  0.2161764  0.2941945
#> vs           0.000000  0.000000  0.00000000  0.000000  0.0000000  0.0000000
#> wt          -5.169043 -5.404898 -5.59195393 -5.730271 -5.6994900 -5.6707821
#>                 1.2351     1.1519     1.0742     1.0018     0.9343     0.8713
#> (Intercept) 31.4400631 30.1448633 28.9369557 27.8104569 26.7598805 25.7801093
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.3669545  0.4348107  0.4980936  0.5571114  0.6121517  0.6634824
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.6440091 -5.6190404 -5.5957546 -5.5740381 -5.5537853 -5.5348974
#>                 0.8126     0.7579     0.7068     0.6591     0.6147    0.5733
#> (Intercept) 24.8663715 24.0142165 23.2194939 22.4783330 21.7871238 21.178539
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> qsec         0.7113536  0.7559984  0.7976343  0.8364641  0.8726769  0.904561
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> wt          -5.5172826 -5.5008549 -5.4855343 -5.4712464 -5.4579213 -5.446189
#>                 0.5347     0.4986     0.4650     0.4337     0.4044     0.3772
#> (Intercept) 21.1785303 21.1785301 21.1785301 21.1785301 21.1785301 21.1785301
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.9045615  0.9045615  0.9045615  0.9045615  0.9045615  0.9045615
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.4461890 -5.4461890 -5.4461890 -5.4461890 -5.4461890 -5.4461890
#>                 0.3518     0.3281     0.3059     0.2853     0.2661     0.2482
#> (Intercept) 21.1785301 21.1785301 21.1785301 21.1785301 21.1785301 29.9595541
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 -0.8193204
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.9045615  0.9045615  0.9045615  0.9045615  0.9045615  0.4905867
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.4461890 -5.4461890 -5.4461890 -5.4461890 -5.4461890 -4.2713680
#>                 0.2314     0.2158     0.2013     0.1877      0.1751     0.1633
#> (Intercept) 29.9617515 29.9635384 29.9649915 29.9661732 30.82610527 31.8586104
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.00000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000 -0.05717182 -0.1246102
#> cyl         -0.8195339 -0.8197075 -0.8198487 -0.8199635 -0.84933701 -0.8857343
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.00000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.00000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.00000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.00000000  0.0000000
#> qsec         0.4904836  0.4903998  0.4903316  0.4902761  0.44695452  0.3950009
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.00000000  0.0000000
#> wt          -4.2710601 -4.2708098 -4.2706062 -4.2704407 -4.19273964 -4.0985662
#>                 0.1523     0.1420     0.1324     0.1235     0.1152      0.1074
#> (Intercept) 32.8225349 33.7214871 33.9299291 33.9333451 33.9362832 31.83406454
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.00000000
#> carb        -0.1875500 -0.2462475 -0.2594969 -0.2595703 -0.2596334  0.81489464
#> cyl         -0.9197321 -0.9514382 -0.9591232 -0.9593849 -0.9596101 -1.27836859
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.02973862
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.00000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.00000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 -0.02853574
#> qsec         0.3464991  0.3012664  0.2907965  0.2906324  0.2904912  0.62803537
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.00000000
#> wt          -4.0106338 -3.9286284 -3.9093579 -3.9089352 -3.9085717 -6.25681545
#>                  0.1002      0.0934      0.0871      0.0813      0.0758
#> (Intercept) 31.83342958 31.83282817 31.83225854 31.83171902 31.83120802
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.81501842  0.81513560  0.81524653  0.81535155  0.81545097
#> cyl         -1.27845955 -1.27854561 -1.27862705 -1.27870411 -1.27877703
#> disp         0.02974312  0.02974738  0.02975141  0.02975522  0.02975884
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.02853758 -0.02853932 -0.02854097 -0.02854254 -0.02854402
#> qsec         0.62810631  0.62817347  0.62823707  0.62829729  0.62835430
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -6.25718210 -6.25752921 -6.25785783 -6.25816894 -6.25846348
#>                  0.0707      0.0659      0.0615      0.0573      0.0535
#> (Intercept) 31.83072404 31.83026566 31.82983154 31.82942039 31.82903101
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.81554508  0.81563418  0.81571853  0.81579839  0.81587398
#> cyl         -1.27884603 -1.27891134 -1.27897314 -1.27903162 -1.27908698
#> disp         0.02976226  0.02976549  0.02976856  0.02977146  0.02977420
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.02854542 -0.02854674 -0.02854800 -0.02854918 -0.02855031
#> qsec         0.62840828  0.62845939  0.62850779  0.62855361  0.62859700
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -6.25874232 -6.25900631 -6.25925624 -6.25949286 -6.25971687
#>                  0.0499      0.0465      0.0434      0.0404      0.0377
#> (Intercept) 31.82866225 31.82831302 31.82798230 31.82766911 31.82737253
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.81594555  0.81601330  0.81607745  0.81613817  0.81619566
#> cyl         -1.27913937 -1.27918895 -1.27923589 -1.27928031 -1.27932235
#> disp         0.02977680  0.02977926  0.02978159  0.02978380  0.02978589
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.02855137 -0.02855238 -0.02855334 -0.02855424 -0.02855509
#> qsec         0.62863807  0.62867697  0.62871379  0.62874866  0.62878167
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -6.25992895 -6.26012974 -6.26031984 -6.26049981 -6.26067020
#>                  0.0352      0.0328      0.0306      0.0285      0.0266
#> (Intercept) 31.82709168 31.82682572 31.82657389 33.02949908 33.03462987
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.81625008  0.81630160  0.81635037  0.83254492  0.83240500
#> cyl         -1.27936215 -1.27939982 -1.27943548 -1.33039660 -1.33066313
#> disp         0.02978786  0.02978974  0.02979151  0.02985776  0.02985441
#> drat         0.00000000  0.00000000  0.00000000 -0.18974718 -0.19010573
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.02855590 -0.02855667 -0.02855740 -0.02828411 -0.02828017
#> qsec         0.62881292  0.62884252  0.62887054  0.61560166  0.61541208
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -6.26083151 -6.26098423 -6.26112882 -6.28467422 -6.28412264
#>                  0.0248      0.0231      0.0216      0.0201      0.0188
#> (Intercept) 33.03937472 33.04376033 33.04780970 33.05154343 32.52857183
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.83226299  0.83212036  0.83197815  0.83183715  0.85035149
#> cyl         -1.33089755 -1.33110273 -1.33128123 -1.33143538 -1.25131048
#> disp         0.02985090  0.02984726  0.02984351  0.02983970  0.03011202
#> drat        -0.19041995 -0.19069527 -0.19093602 -0.19114577 -0.14506707
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.02827648 -0.02827303 -0.02826979 -0.02826675 -0.02905971
#> qsec         0.61522878  0.61505211  0.61488230  0.61471950  0.60258265
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.29440813
#> wt          -6.28357179 -6.28302491 -6.28248465 -6.28195318 -6.30089566
#>                  0.0175      0.0163      0.0152      0.0142      0.0132
#> (Intercept) 32.52805352 32.52760154 32.52721101 32.52687734 32.02828688
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.85048581  0.85061395  0.85073612  0.85085253  0.80430811
#> cyl         -1.25137867 -1.25144856 -1.25151979 -1.25159202 -1.22298414
#> disp         0.03011645  0.03012070  0.03012475  0.03012863  0.02952602
#> drat        -0.14513715 -0.14520828 -0.14528015 -0.14535248 -0.19482860
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.14688068
#> hp          -0.02906157 -0.02906331 -0.02906494 -0.02906648 -0.02926424
#> qsec         0.60266035  0.60273438  0.60280488  0.60287196  0.59873377
#> vs           0.29443640  0.29445519  0.29446534  0.29446762  0.24213360
#> wt          -6.30131119 -6.30170728 -6.30208459 -6.30244378 -6.19446684
#>                  0.0124      0.0115      0.0107      0.0100      0.0093
#> (Intercept) 32.02590217 32.02358373 32.02133083 32.01914267 32.01701838
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.80411242  0.80392515  0.80374601  0.80357470  0.80341093
#> cyl         -1.22270548 -1.22243554 -1.22217417 -1.22192118 -1.22167641
#> disp         0.02952156  0.02951729  0.02951322  0.02950932  0.02950560
#> drat        -0.19479092 -0.19475330 -0.19471578 -0.19467846 -0.19464137
#> gear         0.14726899  0.14764263  0.14800202  0.14834761  0.14867983
#> hp          -0.02926481 -0.02926539 -0.02926597 -0.02926656 -0.02926715
#> qsec         0.59868087  0.59863060  0.59858286  0.59853754  0.59849453
#> vs           0.24227350  0.24241244  0.24255021  0.24268664  0.24282156
#> wt          -6.19396589 -6.19348731 -6.19303027 -6.19259392 -6.19217749
#>                  0.0087      0.0081      0.0076        0.0071      0.0066
#> (Intercept) 32.01495702 32.01295761 32.01101911 32.0069929757 31.62603782
#> am           0.00000000  0.00000000  0.00000000  0.0004874083  0.08434122
#> carb         0.80325443  0.80310492  0.80296215  0.8026776100  0.80666889
#> cyl         -1.22143967 -1.22121080 -1.22098961 -1.2205003343 -1.19357075
#> disp         0.02950206  0.02949867  0.02949544  0.0294896307  0.02952760
#> drat        -0.19460458 -0.19456814 -0.19453208 -0.1946221867 -0.19409385
#> gear         0.14899908  0.14930578  0.14960034  0.1500841214  0.14250648
#> hp          -0.02926774 -0.02926833 -0.02926891 -0.0292713970 -0.02958494
#> qsec         0.59845373  0.59841504  0.59837837  0.5983407987  0.60814301
#> vs           0.24295484  0.24308634  0.24321594  0.2434625978  0.28105757
#> wt          -6.19178019 -6.19140129 -6.19104005 -6.1903085198 -6.17805853
#>                  0.0061      0.0057      0.0053      0.0050      0.0047
#> (Intercept) 31.62213699 31.61841007 31.61485039 31.61145153 31.60820727
#> am           0.08473149  0.08510930  0.08547498  0.08582884  0.08617120
#> carb         0.80679954  0.80693130  0.80706393  0.80719719  0.80733088
#> cyl         -1.19329196 -1.19302820 -1.19277881 -1.19254318 -1.19232070
#> disp         0.02953068  0.02953378  0.02953690  0.02954003  0.02954317
#> drat        -0.19395225 -0.19381570 -0.19368401 -0.19355700 -0.19343450
#> gear         0.14243889  0.14236419  0.14228289  0.14219549  0.14210250
#> hp          -0.02958895 -0.02959282 -0.02959658 -0.02960021 -0.02960373
#> qsec         0.60825189  0.60835881  0.60846375  0.60856672  0.60866772
#> vs           0.28163777  0.28219763  0.28273777  0.28325881  0.28376133
#> wt          -6.17835492 -6.17865529 -6.17895902 -6.17926552 -6.17957422
#> 
#> $iter
#>   [1]   0   6   4   4   3   3   3   2   2   2   2   2   2   2   2   5   4   3
#>  [19]   3   3   3   3   3   3   3   3   3   3   3   3   3   1   1   1   1   1
#>  [37]   1   1   1   1   1   1  42   1   1   1   1  37  38  37  36  15   1   1
#>  [55] 176   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [73]   1  80   1   1   1   1   1  74   1   1   1   1  59   1   1   1   1   1
#>  [91]   1   1   1   2  40   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.650114832 4.336712656 4.044432738 3.771851507 3.517641340 3.280564086
#>   [7] 3.059465047 2.853267342 2.660966673 2.481626425 2.314373110 2.158392108
#>  [13] 2.012923704 1.877259383 1.750738384 1.632744476 1.522702963 1.420077880
#>  [19] 1.324369384 1.235111320 1.151868950 1.074236838 1.001836871 0.934316419
#>  [25] 0.871346619 0.812620773 0.757852853 0.706776107 0.659141763 0.614717814
#>  [31] 0.573287891 0.534650206 0.498616571 0.465011483 0.433671266 0.404443274
#>  [37] 0.377185151 0.351764134 0.328056409 0.305946505 0.285326734 0.266096667
#>  [43] 0.248162643 0.231437311 0.215839211 0.201292370 0.187725938 0.175073838
#>  [49] 0.163274448 0.152270296 0.142007788 0.132436938 0.123511132 0.115186895
#>  [55] 0.107423684 0.100183687 0.093431642 0.087134662 0.081262077 0.075785285
#>  [61] 0.070677611 0.065914176 0.061471781 0.057328789 0.053465021 0.049861657
#>  [67] 0.046501148 0.043367127 0.040444327 0.037718515 0.035176413 0.032805641
#>  [73] 0.030594650 0.028532673 0.026609667 0.024816264 0.023143731 0.021583921
#>  [79] 0.020129237 0.018772594 0.017507384 0.016327445 0.015227030 0.014200779
#>  [85] 0.013243694 0.012351113 0.011518690 0.010742368 0.010018369 0.009343164
#>  [91] 0.008713466 0.008126208 0.007578529 0.007067761 0.006591418 0.006147178
#>  [97] 0.005732879 0.005346502 0.004986166 0.004650115
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 42
#> 
#> $loss
#>   [1] 614.57238 527.39988 454.46713 393.72491 343.40402 301.97795 268.13040
#>   [8] 240.72714 218.79160 201.48367 188.08127 177.96441 170.60138 165.53668
#>  [15] 162.38071 153.05849 143.88207 134.75750 127.18948 120.95091 115.84546
#>  [22] 111.70394 108.38065 105.75022 103.70489 102.15211 101.01247 100.21795
#>  [29]  99.71035  99.43993  99.36399  99.36399  99.36399  99.36399  99.36399
#>  [36]  99.36399  99.36399  99.36399  99.36399  99.36399  99.36399  99.36399
#>  [43]  91.59242  91.59242  91.59241  91.59241  91.59241  90.95062  90.39735
#>  [50]  90.08001  89.95723  89.95262  89.95261  89.95260  78.90651  78.90650
#>  [57]  78.90649  78.90648  78.90648  78.90647  78.90646  78.90646  78.90646
#>  [64]  78.90645  78.90645  78.90645  78.90644  78.90644  78.90644  78.90644
#>  [71]  78.90643  78.90643  78.90643  78.85057  78.85055  78.85054  78.85052
#>  [78]  78.85051  78.85049  78.79111  78.79110  78.79109  78.79108  78.79108
#>  [85]  78.75936  78.75934  78.75933  78.75931  78.75930  78.75928  78.75927
#>  [92]  78.75926  78.75925  78.75916  78.75227  78.75225  78.75224  78.75222
#>  [99]  78.75221  78.75220
#> 
#> $linear.predictors
#>      4.6501   4.3367   4.0444   3.7719   3.5176   3.2806   3.0595   2.8533
#> 1  19.84762 20.19696 20.52276 20.82660 21.10996 21.37423 21.62068 21.85053
#> 2  19.84762 20.37223 20.86149 21.31777 21.74330 22.14015 22.51025 22.85541
#> 3  19.84762 19.84934 19.85095 19.85245 19.85385 19.85516 19.85637 19.85751
#> 4  19.84762 19.71789 19.59691 19.48408 19.37885 19.28072 19.18920 19.10385
#> 5  19.84762 19.70621 19.57433 19.45133 19.33663 19.22966 19.12989 19.03685
#> 6  19.84762 19.86395 19.87918 19.89338 19.90663 19.91898 19.93050 19.94125
#> 7  19.84762 19.88732 19.92434 19.95887 19.99107 20.02111 20.04911 20.07523
#> 8  19.84762 19.71789 19.59691 19.48408 19.37885 19.28072 19.18920 19.10385
#> 9  19.84762 19.54846 19.26947 19.00928 18.76663 18.54033 18.32928 18.13246
#> 10 19.84762 19.51925 19.21302 18.92742 18.66107 18.41268 18.18102 17.96498
#> 11 19.84762 18.55878 17.35680 16.23583 15.19041 14.21545 13.30619 12.45822
#> 12 19.84762 20.44234 20.99698 21.51423 21.99663 22.44651 22.86608 23.25736
#> 13 19.84762 19.67115 19.50658 19.35310 19.20996 19.07647 18.95198 18.83588
#> 14 19.84762 19.72081 19.60255 19.49226 19.38941 19.29348 19.20402 19.12059
#> 15 19.84762 19.48420 19.14527 18.82919 18.53441 18.25949 18.00311 17.76400
#> 16 19.84762 19.48128 19.13963 18.82100 18.52385 18.24673 17.98828 17.74725
#> 17 19.84762 20.59716 21.29618 21.94810 22.55607 23.12307 23.65186 24.14501
#> 18 19.84762 20.84371 21.77266 22.63901 23.44696 24.20047 24.90318 25.55854
#> 19 19.84762 19.87563 19.90176 19.92613 19.94885 19.97004 19.98981 20.00824
#> 20 19.84762 19.64194 19.45013 19.27124 19.10441 18.94882 18.80372 18.66840
#> 21 19.84762 20.10348 20.34211 20.56464 20.77218 20.96574 21.14625 21.31459
#>      2.6610   2.4816   2.3144    2.1584    2.0129    1.8773    1.7507    1.6327
#> 1  22.06488 22.26479 22.45122 22.625088 22.787239 22.938461 23.079491 23.105298
#> 2  23.17731 23.47751 23.75748 24.018578 24.262082 24.489174 24.700960 24.911055
#> 3  19.85857 19.85955 19.86047 19.861333 19.862133 19.862880 19.863576 19.955736
#> 4  19.02425 18.95001 18.88078 18.816215 18.756001 18.699845 18.647474 18.553280
#> 5  18.95009 18.86916 18.79370 18.723316 18.657678 18.596464 18.539376 18.632206
#> 6  19.95127 19.96061 19.96933 19.977457 19.985037 19.992106 19.998699 20.128918
#> 7  20.09959 20.12231 20.14350 20.163256 20.181683 20.198868 20.214895 20.525477
#> 8  19.02425 18.95001 18.88078 18.816215 18.756001 18.699845 18.647474 18.629586
#> 9  17.94890 17.77771 17.61806 17.469175 17.330320 17.200823 17.080054 16.966189
#> 10 17.76350 17.57559 17.40035 17.236926 17.084512 16.942371 16.809809 16.710437
#> 11 11.66740 10.92988 10.24206  9.600601  9.002373  8.444464  7.924156  7.506535
#> 12 23.62228 23.96260 24.27998 24.575974 24.852019 25.109459 25.349548 25.633357
#> 13 18.72760 18.62662 18.53245 18.444618 18.362710 18.286322 18.215082 18.096981
#> 14 19.04279 18.97022 18.90255 18.839440 18.780582 18.725691 18.674499 18.597932
#> 15 17.54101 17.33305 17.13910 16.958228 16.789544 16.632228 16.485515 16.220520
#> 16 17.52247 17.31284 17.11733 16.935004 16.764963 16.606383 16.458490 16.290327
#> 17 24.60492 25.03383 25.43384 25.806891 26.154797 26.479255 26.781846 27.081245
#> 18 26.16973 26.73973 27.27131 27.767067 28.229409 28.660591 29.062713 29.321821
#> 19 20.02543 20.04146 20.05641 20.070356 20.083360 20.095487 20.106797 19.912880
#> 20 18.54220 18.42450 18.31474 18.212369 18.116903 18.027870 17.944837 17.682060
#> 21 21.47159 21.61800 21.75455 21.881893 22.000656 22.111414 22.214707 22.338160
#>       1.5227    1.4201    1.3244    1.2351    1.1519   1.0742   1.0018
#> 1  23.082770 22.943628 22.813856 22.692830 22.579961 22.47470 22.37653
#> 2  25.086780 25.118254 25.147609 25.174985 25.200516 25.22433 25.24653
#> 3  20.068184 20.196637 20.316440 20.428169 20.532368 20.62954 20.72017
#> 4  18.458162 18.391105 18.328564 18.270237 18.215842 18.16511 18.11780
#> 5  18.767637 18.968880 19.156570 19.331611 19.494855 19.64710 19.78908
#> 6  20.285654 20.460183 20.622959 20.774764 20.916338 21.04837 21.17150
#> 7  20.899188 21.315074 21.702954 22.064692 22.402050 22.71667 23.01009
#> 8  18.627794 18.667811 18.705133 18.739939 18.772399 18.80267 18.83090
#> 9  16.873248 16.863635 16.854670 16.846308 16.838510 16.83124 16.82446
#> 10 16.639745 16.665131 16.688808 16.710889 16.731482 16.75069 16.76860
#> 11  7.195325  7.256258  7.313087  7.366087  7.415514  7.46161  7.50460
#> 12 25.888384 25.988105 26.081110 26.167847 26.248738 26.32418 26.39453
#> 13 17.979862 17.902719 17.830772 17.763673 17.701097 17.64274 17.58831
#> 14 18.523921 18.480132 18.439292 18.401204 18.365684 18.33256 18.30166
#> 15 15.952689 15.763265 15.586598 15.421837 15.268180 15.12488 14.99124
#> 16 16.141378 16.089297 16.040723 15.995422 15.953175 15.91377 15.87703
#> 17 27.331367 27.375249 27.416176 27.454345 27.489942 27.52314 27.55410
#> 18 29.484491 29.348081 29.220857 29.102208 28.991555 28.88836 28.79212
#> 19 19.671373 19.385203 19.118305 18.869394 18.637260 18.42077 18.21887
#> 20 17.392517 17.127024 16.879411 16.648486 16.433125 16.23228 16.04497
#> 21 22.449530 22.494327 22.536107 22.575071 22.611409 22.64530 22.67690
#>       0.9343    0.8713    0.8126    0.7579   0.7068    0.6591    0.6147
#> 1  22.284979 22.199598 22.119971 22.045711 21.97646 21.911867 21.851632
#> 2  25.267241 25.286554 25.304566 25.321364 25.33703 25.351639 25.365264
#> 3  20.804689 20.883511 20.957021 21.025577 21.08951 21.149139 21.204746
#> 4  18.073680 18.032532 17.994157 17.958369 17.92499 17.893865 17.864836
#> 5  19.921490 20.044978 20.160143 20.267547 20.36771 20.461125 20.548244
#> 6  21.286339 21.393434 21.493312 21.586458 21.67333 21.754340 21.829894
#> 7  23.283730 23.538929 23.776928 23.998887 24.20589 24.398936 24.578974
#> 8  18.857234 18.881790 18.904690 18.926047 18.94596 18.964539 18.981862
#> 9  16.818130 16.812232 16.806730 16.801600 16.79682 16.792353 16.788191
#> 10 16.785302 16.800880 16.815408 16.828957 16.84159 16.853376 16.864366
#> 11  7.544692  7.582082  7.616952  7.649471  7.67980  7.708084  7.734462
#> 12 26.460146 26.521337 26.578404 26.631625 26.68126 26.727548 26.770717
#> 13 17.537555 17.490218 17.446072 17.404901 17.36650 17.330696 17.297301
#> 14 18.272852 18.245982 18.220923 18.197553 18.17576 18.155431 18.136475
#> 15 14.866602 14.750367 14.641965 14.540869 14.44659 14.358659 14.276658
#> 16 15.842762 15.810803 15.780998 15.753202 15.72728 15.703104 15.680558
#> 17 27.582972 27.609900 27.635012 27.658432 27.68027 27.700644 27.719640
#> 18 28.702366 28.618662 28.540598 28.467796 28.39990 28.336581 28.277529
#> 19 18.030580 17.854979 17.691213 17.538484 17.39605 17.263212 17.139329
#> 20 15.870281 15.707368 15.555435 15.413741 15.28160 15.158360 15.043428
#> 21 22.706378 22.733867 22.759502 22.783410 22.80571 22.826501 22.845894
#>       0.5733    0.5347    0.4986    0.4650    0.4337    0.4044    0.3772
#> 1  21.798598 21.798597 21.798597 21.798597 21.798597 21.798597 21.798597
#> 2  25.377261 25.377261 25.377261 25.377261 25.377261 25.377261 25.377261
#> 3  21.253707 21.253708 21.253708 21.253708 21.253708 21.253708 21.253708
#> 4  17.839277 17.839276 17.839276 17.839276 17.839276 17.839276 17.839276
#> 5  20.624948 20.624949 20.624949 20.624949 20.624949 20.624949 20.624949
#> 6  21.896416 21.896417 21.896417 21.896417 21.896417 21.896417 21.896417
#> 7  24.737491 24.737493 24.737493 24.737493 24.737493 24.737493 24.737493
#> 8  18.997115 18.997115 18.997115 18.997115 18.997115 18.997115 18.997115
#> 9  16.784527 16.784527 16.784527 16.784527 16.784527 16.784527 16.784527
#> 10 16.874042 16.874042 16.874042 16.874042 16.874042 16.874042 16.874042
#> 11  7.757686  7.757687  7.757687  7.757687  7.757687  7.757687  7.757687
#> 12 26.808726 26.808726 26.808726 26.808726 26.808726 26.808726 26.808726
#> 13 17.267898 17.267897 17.267897 17.267897 17.267897 17.267897 17.267897
#> 14 18.119785 18.119785 18.119785 18.119785 18.119785 18.119785 18.119785
#> 15 14.204458 14.204457 14.204457 14.204457 14.204457 14.204457 14.204457
#> 16 15.660707 15.660707 15.660707 15.660707 15.660707 15.660707 15.660707
#> 17 27.736366 27.736366 27.736366 27.736366 27.736366 27.736366 27.736366
#> 18 28.225536 28.225535 28.225535 28.225535 28.225535 28.225535 28.225535
#> 19 17.030254 17.030253 17.030253 17.030253 17.030253 17.030253 17.030253
#> 20 14.942234 14.942233 14.942233 14.942233 14.942233 14.942233 14.942233
#> 21 22.862968 22.862968 22.862968 22.862968 22.862968 22.862968 22.862968
#>       0.3518    0.3281    0.3059    0.2853    0.2661    0.2482    0.2314
#> 1  21.798597 21.798597 21.798597 21.798597 21.798597 21.927705 21.927731
#> 2  25.377261 25.377261 25.377261 25.377261 25.377261 25.902518 25.902656
#> 3  21.253708 21.253708 21.253708 21.253708 21.253708 20.848190 20.848091
#> 4  17.839276 17.839276 17.839276 17.839276 17.839276 17.061271 17.061065
#> 5  20.624949 20.624949 20.624949 20.624949 20.624949 20.184362 20.184259
#> 6  21.896417 21.896417 21.896417 21.896417 21.896417 22.868343 22.868606
#> 7  24.737493 24.737493 24.737493 24.737493 24.737493 24.461899 24.461851
#> 8  18.997115 18.997115 18.997115 18.997115 18.997115 19.327863 19.327951
#> 9  16.784527 16.784527 16.784527 16.784527 16.784527 16.107115 16.106938
#> 10 16.874042 16.874042 16.874042 16.874042 16.874042 16.089781 16.089578
#> 11  7.757687  7.757687  7.757687  7.757687  7.757687  8.979346  8.979668
#> 12 26.808726 26.808726 26.808726 26.808726 26.808726 26.836987 26.837000
#> 13 17.267897 17.267897 17.267897 17.267897 17.267897 16.645974 16.645807
#> 14 18.119785 18.119785 18.119785 18.119785 18.119785 17.219992 17.219755
#> 15 14.204457 14.204457 14.204457 14.204457 14.204457 14.562879 14.562962
#> 16 15.660707 15.660707 15.660707 15.660707 15.660707 15.346085 15.346000
#> 17 27.736366 27.736366 27.736366 27.736366 27.736366 27.689265 27.689255
#> 18 28.225535 28.225535 28.225535 28.225535 28.225535 28.510608 28.510675
#> 19 17.030253 17.030253 17.030253 17.030253 17.030253 16.978262 16.978232
#> 20 14.942233 14.942233 14.942233 14.942233 14.942233 15.318773 15.318856
#> 21 22.862968 22.862968 22.862968 22.862968 22.862968 23.932783 23.933064
#>      0.2158    0.2013    0.1877    0.1751    0.1633    0.1523    0.1420
#> 1  21.92775 21.927769 21.927783 21.873289 21.809235 21.749456 21.693707
#> 2  25.90277 25.902861 25.902935 25.962253 26.033356 26.099734 26.161638
#> 3  20.84801 20.847946 20.847893 20.882049 20.921521 20.958346 20.992690
#> 4  17.06090 17.060760 17.060649 17.101207 17.147363 17.190412 17.230560
#> 5  20.18417 20.184106 20.184051 20.203453 20.225473 20.246010 20.265164
#> 6  22.86882 22.868994 22.869136 22.878665 22.892044 22.904566 22.916244
#> 7  24.46181 24.461780 24.461754 24.342542 24.201489 24.069839 23.947061
#> 8  19.32802 19.328082 19.328129 19.257639 19.175212 19.098295 19.026562
#> 9  16.10679 16.106677 16.106581 16.087374 16.063269 16.040748 16.019745
#> 10 16.08941 16.089279 16.089170 16.056519 16.016341 15.978816 15.943820
#> 11  8.97993  8.980143  8.980316  9.026032  9.082588  9.135414  9.184679
#> 12 26.83701 26.837019 26.837026 26.849763 26.864884 26.878999 26.892162
#> 13 16.64567 16.645562 16.645472 16.698745 16.760227 16.817586 16.871080
#> 14 17.21956 17.219406 17.219278 17.247318 17.278456 17.307485 17.334558
#> 15 14.56303 14.563084 14.563128 14.590171 14.622764 14.653195 14.681574
#> 16 15.34593 15.345874 15.345829 15.416556 15.499293 15.576500 15.648503
#> 17 27.68925 27.689240 27.689235 27.706075 27.725854 27.744312 27.761527
#> 18 28.51073 28.510773 28.510809 28.524330 28.540837 28.556252 28.570628
#> 19 16.97821 16.978188 16.978172 16.992578 17.009353 17.025006 17.039603
#> 20 15.31892 15.318979 15.319024 15.131490 14.910986 14.705202 14.513288
#> 21 23.93329 23.933479 23.933630 23.971951 24.019455 24.063827 24.105208
#>       0.1324    0.1235    0.1152    0.1074   0.1002    0.0934   0.0871
#> 1  21.681195 21.681153 21.681116 22.987285 22.98732 22.987360 22.98740
#> 2  26.175952 26.176173 26.176364 25.265358 25.26527 25.265181 25.26510
#> 3  21.000191 21.000131 21.000080 21.605726 21.60583 21.605923 21.60601
#> 4  17.239115 17.238951 17.238809 18.114770 18.11490 18.115020 18.11514
#> 5  20.269220 20.269135 20.269063 19.723978 19.72391 19.723837 19.72377
#> 6  22.919521 22.919808 22.920056 23.545285 23.54533 23.545370 23.54541
#> 7  23.919205 23.919000 23.918823 24.499723 24.49990 24.500067 24.50023
#> 8  19.010587 19.010589 19.010591 18.867330 18.86721 18.867094 18.86699
#> 9  16.014566 16.014356 16.014175 14.832778 14.83258 14.832387 14.83221
#> 10 15.935417 15.935162 15.934943 14.771152 14.77096 14.770780 14.77061
#> 11  9.196592  9.196988  9.197329  9.665898  9.66598  9.666057  9.66613
#> 12 26.895160 26.895189 26.895215 26.455410 26.45534 26.455278 26.45522
#> 13 16.882747 16.882641 16.882550 16.984391 16.98434 16.984285 16.98424
#> 14 17.340085 17.339872 17.339689 17.369935 17.36988 17.369827 17.36978
#> 15 14.688196 14.688318 14.688423 13.935808 13.93575 13.935694 13.93564
#> 16 15.664549 15.664551 15.664552 16.789146 16.78931 16.789461 16.78961
#> 17 27.765386 27.765397 27.765406 27.764407 27.76440 27.764388 27.76438
#> 18 28.574045 28.574133 28.574208 29.101220 29.10133 29.101439 29.10154
#> 19 17.042841 17.042829 17.042819 17.043922 17.04401 17.044101 17.04418
#> 20 14.470189 14.470037 14.469906 14.350610 14.35070 14.350788 14.35087
#> 21 24.115242 24.115587 24.115883 23.125868 23.12576 23.125659 23.12556
#>       0.0813    0.0758    0.0707    0.0659   0.0615    0.0573    0.0535
#> 1  22.987428 22.987459 22.987489 22.987517 22.98754 22.987568 22.987592
#> 2  25.265023 25.264950 25.264881 25.264815 25.26475 25.264695 25.264639
#> 3  21.606101 21.606182 21.606260 21.606333 21.60640 21.606467 21.606530
#> 4  18.115245 18.115348 18.115446 18.115538 18.11563 18.115709 18.115787
#> 5  19.723712 19.723654 19.723599 19.723547 19.72350 19.723451 19.723407
#> 6  23.545447 23.545482 23.545515 23.545546 23.54558 23.545604 23.545631
#> 7  24.500376 24.500518 24.500653 24.500780 24.50090 24.501015 24.501123
#> 8  18.866883 18.866786 18.866694 18.866607 18.86652 18.866446 18.866373
#> 9  14.832037 14.831876 14.831723 14.831579 14.83144 14.831313 14.831190
#> 10 14.770448 14.770294 14.770149 14.770012 14.76988 14.769759 14.769643
#> 11  9.666199  9.666265  9.666327  9.666385  9.66644  9.666493  9.666543
#> 12 26.455159 26.455105 26.455053 26.455004 26.45496 26.454914 26.454873
#> 13 16.984190 16.984147 16.984106 16.984067 16.98403 16.983995 16.983961
#> 14 17.369730 17.369685 17.369643 17.369603 17.36956 17.369529 17.369495
#> 15 13.935592 13.935544 13.935500 13.935457 13.93542 13.935379 13.935344
#> 16 16.789744 16.789874 16.789997 16.790114 16.79022 16.790329 16.790428
#> 17 27.764371 27.764363 27.764356 27.764349 27.76434 27.764336 27.764330
#> 18 29.101635 29.101726 29.101811 29.101892 29.10197 29.102042 29.102110
#> 19 17.044261 17.044335 17.044405 17.044471 17.04453 17.044593 17.044649
#> 20 14.350948 14.351022 14.351092 14.351158 14.35122 14.351280 14.351336
#> 21 23.125471 23.125385 23.125303 23.125226 23.12515 23.125083 23.125017
#>       0.0499    0.0465    0.0434    0.0404    0.0377    0.0352    0.0328
#> 1  22.987614 22.987636 22.987656 22.987675 22.987693 22.987710 22.987726
#> 2  25.264587 25.264537 25.264490 25.264445 25.264403 25.264363 25.264325
#> 3  21.606588 21.606644 21.606697 21.606746 21.606794 21.606838 21.606881
#> 4  18.115862 18.115932 18.115999 18.116062 18.116121 18.116178 18.116232
#> 5  19.723366 19.723326 19.723289 19.723254 19.723220 19.723189 19.723159
#> 6  23.545656 23.545680 23.545702 23.545724 23.545744 23.545763 23.545781
#> 7  24.501225 24.501322 24.501413 24.501500 24.501582 24.501660 24.501734
#> 8  18.866303 18.866236 18.866174 18.866114 18.866058 18.866005 18.865955
#> 9  14.831074 14.830964 14.830860 14.830762 14.830669 14.830581 14.830497
#> 10 14.769533 14.769429 14.769330 14.769236 14.769148 14.769064 14.768985
#> 11  9.666589  9.666634  9.666676  9.666716  9.666753  9.666789  9.666823
#> 12 26.454834 26.454797 26.454761 26.454728 26.454697 26.454667 26.454638
#> 13 16.983930 16.983900 16.983872 16.983846 16.983821 16.983797 16.983774
#> 14 17.369463 17.369433 17.369404 17.369377 17.369351 17.369327 17.369304
#> 15 13.935310 13.935277 13.935247 13.935218 13.935191 13.935165 13.935140
#> 16 16.790522 16.790610 16.790694 16.790774 16.790849 16.790920 16.790988
#> 17 27.764324 27.764319 27.764314 27.764309 27.764304 27.764300 27.764296
#> 18 29.102176 29.102237 29.102295 29.102351 29.102403 29.102452 29.102499
#> 19 17.044702 17.044752 17.044800 17.044845 17.044887 17.044928 17.044966
#> 20 14.351389 14.351439 14.351486 14.351531 14.351574 14.351614 14.351653
#> 21 23.124955 23.124896 23.124840 23.124787 23.124737 23.124690 23.124645
#>       0.0306    0.0285    0.0266   0.0248    0.0231    0.0216    0.0201
#> 1  22.987742 22.970232 22.970028 22.96984 22.969674 22.969523 22.969386
#> 2  25.264289 25.280050 25.280351 25.28064 25.280907 25.281164 25.281406
#> 3  21.606921 21.649362 21.649307 21.64924 21.649170 21.649091 21.649006
#> 4  18.116282 18.111048 18.110792 18.11054 18.110303 18.110069 18.109842
#> 5  19.723130 19.806619 19.806757 19.80688 19.806998 19.807103 19.807199
#> 6  23.545798 23.563276 23.563459 23.56363 23.563782 23.563924 23.564053
#> 7  24.501804 24.446730 24.446409 24.44609 24.445786 24.445487 24.445197
#> 8  18.865907 18.804937 18.804855 18.80480 18.804766 18.804753 18.804759
#> 9  14.830418 14.837822 14.837807 14.83781 14.837828 14.837860 14.837905
#> 10 14.768910 14.769829 14.769766 14.76972 14.769697 14.769689 14.769695
#> 11  9.666855  9.682699  9.682983  9.68324  9.683473  9.683681  9.683868
#> 12 26.454612 26.408825 26.408806 26.40880 26.408802 26.408815 26.408835
#> 13 16.983753 17.043012 17.043011 17.04301 17.043015 17.043020 17.043026
#> 14 17.369282 17.349908 17.349686 17.34949 17.349312 17.349156 17.349019
#> 15 13.935117 13.882631 13.882723 13.88282 13.882912 13.883008 13.883102
#> 16 16.791051 16.791815 16.791669 16.79152 16.791368 16.791215 16.791060
#> 17 27.764292 27.732328 27.732270 27.73222 27.732178 27.732143 27.732114
#> 18 29.102543 29.195981 29.196171 29.19633 29.196455 29.196556 29.196633
#> 19 17.045002 16.932649 16.932439 16.93224 16.932060 16.931888 16.931726
#> 20 14.351689 14.438488 14.438611 14.43870 14.438762 14.438801 14.438821
#> 21 23.124603 23.101759 23.102102 23.10243 23.102751 23.103056 23.103347
#>       0.0188    0.0175    0.0163    0.0152    0.0142    0.0132    0.0124
#> 1  22.887872 22.887904 22.887936 22.887968 22.887999 22.866057 22.865927
#> 2  25.255112 25.255001 25.254898 25.254801 25.254710 25.258733 25.258748
#> 3  21.747822 21.747955 21.748081 21.748198 21.748309 21.700214 21.700094
#> 4  18.097583 18.097731 18.097872 18.098007 18.098134 18.069570 18.069423
#> 5  19.872140 19.872118 19.872096 19.872073 19.872050 19.883890 19.883989
#> 6  23.550671 23.550682 23.550694 23.550708 23.550724 23.587116 23.587136
#> 7  24.380199 24.380349 24.380493 24.380632 24.380766 24.386488 24.386351
#> 8  18.972472 18.972342 18.972212 18.972083 18.971954 18.970464 18.970746
#> 9  14.801048 14.800878 14.800715 14.800559 14.800409 14.807922 14.808064
#> 10 14.727036 14.726876 14.726723 14.726576 14.726436 14.737692 14.737838
#> 11  9.649949  9.649983  9.650019  9.650057  9.650096  9.678607  9.678546
#> 12 26.398405 26.398315 26.398229 26.398147 26.398068 26.397191 26.397256
#> 13 17.021488 17.021479 17.021470 17.021462 17.021455 17.051699 17.051787
#> 14 17.338030 17.338000 17.337970 17.337940 17.337909 17.346336 17.346436
#> 15 13.888331 13.888242 13.888156 13.888073 13.887994 13.779680 13.779454
#> 16 16.778433 16.778598 16.778755 16.778906 16.779051 16.773451 16.773325
#> 17 27.733703 27.733681 27.733659 27.733638 27.733617 27.706305 27.706265
#> 18 29.201836 29.201974 29.202107 29.202237 29.202363 29.234437 29.234374
#> 19 16.968475 16.968529 16.968577 16.968621 16.968661 17.056926 17.057192
#> 20 14.439587 14.439713 14.439835 14.439954 14.440069 14.434667 14.434502
#> 21 23.089809 23.089651 23.089501 23.089360 23.089227 23.072554 23.072547
#>       0.0115    0.0107    0.0100    0.0093    0.0087    0.0081    0.0076
#> 1  22.865800 22.865678 22.865561 22.865447 22.865337 22.865231 22.865129
#> 2  25.258761 25.258771 25.258780 25.258787 25.258792 25.258796 25.258799
#> 3  21.699981 21.699875 21.699774 21.699679 21.699590 21.699505 21.699426
#> 4  18.069283 18.069149 18.069022 18.068901 18.068786 18.068676 18.068572
#> 5  19.884086 19.884180 19.884270 19.884357 19.884442 19.884523 19.884602
#> 6  23.587155 23.587172 23.587188 23.587202 23.587215 23.587226 23.587237
#> 7  24.386220 24.386094 24.385973 24.385857 24.385745 24.385639 24.385536
#> 8  18.971020 18.971285 18.971541 18.971790 18.972030 18.972262 18.972487
#> 9  14.808200 14.808329 14.808453 14.808571 14.808684 14.808791 14.808894
#> 10 14.737978 14.738111 14.738238 14.738360 14.738476 14.738587 14.738693
#> 11  9.678485  9.678425  9.678367  9.678309  9.678252  9.678197  9.678142
#> 12 26.397317 26.397376 26.397432 26.397485 26.397536 26.397585 26.397631
#> 13 17.051871 17.051953 17.052031 17.052106 17.052178 17.052247 17.052314
#> 14 17.346533 17.346627 17.346718 17.346805 17.346890 17.346972 17.347051
#> 15 13.779236 13.779025 13.778823 13.778627 13.778439 13.778258 13.778084
#> 16 16.773204 16.773088 16.772978 16.772873 16.772773 16.772677 16.772587
#> 17 27.706227 27.706190 27.706155 27.706122 27.706090 27.706059 27.706029
#> 18 29.234314 29.234255 29.234199 29.234146 29.234094 29.234044 29.233997
#> 19 17.057449 17.057698 17.057939 17.058171 17.058396 17.058613 17.058822
#> 20 14.434343 14.434190 14.434043 14.433902 14.433766 14.433635 14.433510
#> 21 23.072538 23.072528 23.072516 23.072503 23.072489 23.072474 23.072458
#>       0.0071    0.0066    0.0061    0.0057    0.0053    0.0050    0.0047
#> 1  22.865067 22.882282 22.882317 22.882354 22.882393 22.882434 22.882476
#> 2  25.258907 25.268613 25.268503 25.268395 25.268289 25.268186 25.268085
#> 3  21.699321 21.705668 21.705902 21.706133 21.706360 21.706583 21.706802
#> 4  18.068410 18.057576 18.057652 18.057729 18.057809 18.057889 18.057970
#> 5  19.884880 19.902019 19.902139 19.902253 19.902361 19.902463 19.902560
#> 6  23.587024 23.552255 23.551999 23.551751 23.551509 23.551276 23.551049
#> 7  24.385117 24.367835 24.367788 24.367747 24.367711 24.367678 24.367650
#> 8  18.972637 18.967897 18.968054 18.968199 18.968334 18.968459 18.968575
#> 9  14.809121 14.806710 14.806603 14.806496 14.806388 14.806281 14.806173
#> 10 14.738942 14.741064 14.740986 14.740906 14.740826 14.740744 14.740661
#> 11  9.678166  9.698636  9.698599  9.698564  9.698532  9.698502  9.698474
#> 12 26.397836 26.421977 26.422046 26.422111 26.422172 26.422230 26.422284
#> 13 17.052557 17.047271 17.047222 17.047173 17.047124 17.047074 17.047024
#> 14 17.347262 17.344824 17.344860 17.344892 17.344921 17.344947 17.344971
#> 15 13.777538 13.741783 13.741598 13.741424 13.741260 13.741106 13.740962
#> 16 16.772494 16.768397 16.768469 16.768543 16.768618 16.768694 16.768770
#> 17 27.706060 27.721379 27.721465 27.721550 27.721633 27.721714 27.721793
#> 18 29.233811 29.206480 29.206354 29.206234 29.206120 29.206010 29.205906
#> 19 17.059178 17.069334 17.069500 17.069654 17.069798 17.069931 17.070055
#> 20 14.433192 14.440673 14.440720 14.440768 14.440820 14.440873 14.440928
#> 21 23.072480 23.087329 23.087225 23.087123 23.087024 23.086927 23.086833
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 22.8 21.4 18.7 18.1 24.4 22.8 19.2 17.3 15.2 10.4 32.4 15.5 15.2 13.3
#> [16] 19.2 27.3 30.4 15.8 15.0 21.4
#> 
#> $X
#>               am       carb        cyl       disp        drat       gear
#>  [1,]  1.2747549  0.8536800 -0.1608169 -0.6856670  0.80514087  0.4677072
#>  [2,]  1.2747549 -1.0008662 -1.2865350 -1.1491449  0.69906836  0.4677072
#>  [3,] -0.7844645 -1.0008662 -0.1608169  0.1878104 -0.93444832 -0.9354143
#>  [4,] -0.7844645 -0.3826841  0.9649013  1.0969399 -0.78594680 -0.9354143
#>  [5,] -0.7844645 -1.0008662 -0.1608169 -0.1063198 -1.61331239 -0.9354143
#>  [6,] -0.7844645 -0.3826841 -1.2865350 -0.8042104  0.35963633  0.4677072
#>  [7,] -0.7844645 -0.3826841 -1.2865350 -0.8567973  0.84756988  0.4677072
#>  [8,] -0.7844645  0.8536800 -0.1608169 -0.6179280  0.84756988  0.4677072
#>  [9,] -0.7844645  0.2354979  0.9649013  0.3464624 -0.95566282 -0.9354143
#> [10,] -0.7844645  0.2354979  0.9649013  0.3464624 -0.95566282 -0.9354143
#> [11,] -0.7844645  0.8536800  0.9649013  1.9882434 -1.10416434 -0.9354143
#> [12,]  1.2747549 -1.0008662 -1.2865350 -1.4102968  1.18700192  0.4677072
#> [13,] -0.7844645 -0.3826841  0.9649013  0.7225925 -1.61331239 -0.9354143
#> [14,] -0.7844645 -0.3826841  0.9649013  0.5978100 -0.78594680 -0.9354143
#> [15,] -0.7844645  0.8536800  0.9649013  1.0078096  0.44449433 -0.9354143
#> [16,] -0.7844645 -0.3826841  0.9649013  1.4534613 -0.93444832 -0.9354143
#> [17,]  1.2747549 -1.0008662 -1.2865350 -1.4076229  1.18700192  0.4677072
#> [18,]  1.2747549 -0.3826841 -1.2865350 -1.2641230  0.52935234  1.8708287
#> [19,]  1.2747549  0.8536800  0.9649013  1.0167226  1.48400495  1.8708287
#> [20,]  1.2747549  3.3264082  0.9649013  0.5710709  0.04141879  1.8708287
#> [21,]  1.2747549 -0.3826841 -1.2865350 -1.0332754  1.25064542  0.4677072
#>                hp        qsec         vs           wt
#>  [1,] -0.55793030 -0.75606397 -0.9534626 -0.743118820
#>  [2,] -0.80352970  0.37017451  1.0488088 -1.115950592
#>  [3,] -0.55793030  0.80495494  1.0488088 -0.003669138
#>  [4,]  0.38112624 -0.46271813 -0.9534626  0.275954691
#>  [5,] -0.63016541  1.21354378  1.0488088  0.300810143
#>  [6,] -1.25138743  1.09830077  1.0488088 -0.034738452
#>  [7,] -0.77463565  2.61741313  1.0488088 -0.084449355
#>  [8,] -0.37011899  0.20778663  1.0488088  0.275954691
#>  [9,]  0.45336136 -0.15889566 -0.9534626  0.636358738
#> [10,]  0.45336136  0.05063708 -0.9534626  0.698497367
#> [11,]  0.95900719 -0.04365265 -0.9534626  2.741615479
#> [12,] -1.19359934  0.82066990  1.0488088 -1.265083301
#> [13,]  0.01995065 -0.54129291 -0.9534626  0.375376497
#> [14,]  0.01995065 -0.31604521 -0.9534626  0.269740828
#> [15,]  1.39241790 -1.30608741 -0.9534626  0.773063721
#> [16,]  0.38112624 -0.44700318 -0.9534626  0.779277584
#> [17,] -1.19359934  0.52208574  1.0488088 -1.594418034
#> [18,] -0.51458922 -0.52557795  1.0488088 -2.118868060
#> [19,]  1.66691135 -1.78277439 -0.9534626 -0.059593904
#> [20,]  2.69265003 -1.73039120 -0.9534626  0.437515126
#> [21,] -0.57237732  0.36493619  1.0488088 -0.544275208
#> attr(,"center")
#>  [1]   0.3809524   2.6190476   6.2857143 236.9285714   3.5204762   3.6666667
#>  [7] 148.6190476  17.9033333   0.4761905   3.2179524
#> attr(,"scale")
#>  [1]   0.4856209   1.6176465   1.7766436 112.1952301   0.4713757   0.7126966
#>  [7]  69.2184095   1.9090095   0.4994328   0.8046525
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 34
#> 
#> $lambda.min
#> [1] 0.4650115
#> 
#> $null.dev
#> [1] 29.26535
#> 
#> $Bias
#> [1] 3.584045
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 9.394952