Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 44.31035 41.60154 37.27137 32.94876 29.31816 26.36800 23.99565 22.11282
#>   [9] 20.64357 19.52270 18.69423 18.11023 17.72964 17.51738 17.46294 17.56285
#>  [17] 17.62296 17.32396 17.14966 17.15119 17.01791 16.77326 16.54595 16.35500
#>  [25] 16.15128 15.71791 15.31420 14.94782 14.63432 14.21327 11.80044 11.29815
#>  [33] 10.89277 11.04010 10.98086 10.96931 10.90133 11.00143 11.31445 12.98894
#>  [41] 12.98958 12.83479 12.03723 12.02411 11.69214 11.93055 15.89734 16.81439
#>  [49] 16.57968 21.60776 24.86911 24.87046 27.89203 27.89237 28.09309 27.94265
#>  [57] 27.75573 30.15374 30.15400 30.15426 30.16343 30.69079 30.69110 28.47783
#>  [65] 28.47713 28.41751 28.41698 28.41648 28.36280 28.20148 27.85228 27.54732
#>  [73] 27.54320 27.59871 27.59739 27.81711 27.81571 27.81178 27.81044 27.61788
#>  [81] 27.39824 27.39766 27.37017 27.37014 27.30862 27.30870 27.30881 27.30893
#>  [89] 27.30908 27.30924 27.30943 27.30963 27.30986 27.23044 27.23113 27.23182
#>  [97] 27.23253 27.23366 27.23849 27.23911
#> 
#> $cvse
#>   [1] 12.975197 12.462700 11.506673 10.325885  9.288357  8.407488  7.649250
#>   [8]  6.987622  6.403552  5.883831  5.420008  5.007395  4.644175  4.330582
#>  [15]  4.066179  3.851664  3.705198  3.600583  3.536944  3.541722  3.539780
#>  [22]  3.524965  3.529560  3.513467  3.452750  3.389111  3.338442  3.300495
#>  [29]  3.269163  3.133926  2.404538  2.354069  2.324989  2.544963  2.608769
#>  [36]  2.638611  2.487123  2.483641  2.364629  3.123680  3.124154  3.151714
#>  [43]  3.236468  3.239099  3.240144  3.239321  6.639179  6.652876  6.658079
#>  [50]  6.910433  7.400242  7.400826  9.092717  9.092549  9.065314  9.056023
#>  [57]  9.070985  9.986276  9.986308  9.986338  9.985054  9.933221  9.933250
#>  [64]  8.301354  8.300776  8.314471  8.313960  8.313481  8.310777  8.314399
#>  [71]  8.120280  8.134622  8.134479  8.167171  8.166622  8.162609  8.162093
#>  [78]  8.161997  8.161509  8.104224  8.119116  8.118769  8.119672  8.119334
#>  [85]  8.123113  8.122805  8.122511  8.122232  8.121968  8.121717  8.121480
#>  [92]  8.121256  8.121045  8.122532  8.122342  8.122163  8.121997  8.121810
#>  [99]  8.120776  8.120641
#> 
#> $fold
#>  [1]  8  3  1  8  2  2  7  5  1  5  1  9  3  4  6  7 10  9  4 10  6
#> 
#> $lambda
#>   [1] 5.441598418 5.074852897 4.732824797 4.413848246 4.116369647 3.838940110
#>   [7] 3.580208395 3.338914331 3.113882680 2.904017410 2.708296356 2.525766246
#>  [13] 2.355538054 2.196782673 2.048726873 1.910649539 1.781878155 1.661785531
#>  [19] 1.549786748 1.445336308 1.347925478 1.257079812 1.172356840 1.093343913
#>  [25] 1.019656192 0.950934778 0.886844956 0.827074575 0.771332516 0.719347286
#>  [31] 0.670865687 0.625651585 0.583484762 0.544159842 0.507485290 0.473282480
#>  [37] 0.441384825 0.411636965 0.383894011 0.358020839 0.333891433 0.311388268
#>  [43] 0.290401741 0.270829636 0.252576625 0.235553805 0.219678267 0.204872687
#>  [49] 0.191064954 0.178187815 0.166178553 0.154978675 0.144533631 0.134792548
#>  [55] 0.125707981 0.117235684 0.109334391 0.101965619 0.095093478 0.088684496
#>  [61] 0.082707457 0.077133252 0.071934729 0.067086569 0.062565158 0.058348476
#>  [67] 0.054415984 0.050748529 0.047328248 0.044138482 0.041163696 0.038389401
#>  [73] 0.035802084 0.033389143 0.031138827 0.029040174 0.027082964 0.025257662
#>  [79] 0.023555381 0.021967827 0.020487269 0.019106495 0.017818782 0.016617855
#>  [85] 0.015497867 0.014453363 0.013479255 0.012570798 0.011723568 0.010933439
#>  [91] 0.010196562 0.009509348 0.008868450 0.008270746 0.007713325 0.007193473
#>  [97] 0.006708657 0.006256516 0.005834848 0.005441598
#> 
#> $fit
#> $beta
#>              5.4416     5.0749    4.7328    4.4138    4.1164    3.8389
#> (Intercept) 20.5381 22.3591532 24.057478 25.641341 27.118458 28.496021
#> am           0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.0000 -0.5763363 -1.113829 -1.615097 -2.082581 -2.518559
#>                3.5802    3.3389    3.1139    2.9040    2.7083    2.5258
#> (Intercept) 29.780742 30.978876 32.096260 33.138337 34.110181 35.016525
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.925153 -3.304343 -3.657978 -3.987779 -4.295352 -4.582196
#>                2.3555   2.1968    2.0487    1.9106    1.7819      1.6618
#> (Intercept) 35.861786 36.65008 37.385242 38.070859 38.551433 37.65481476
#> am           0.000000  0.00000  0.000000  0.000000  0.000000  0.00000000
#> carb         0.000000  0.00000  0.000000  0.000000  0.000000  0.00000000
#> cyl          0.000000  0.00000  0.000000  0.000000  0.000000  0.00000000
#> disp         0.000000  0.00000  0.000000  0.000000  0.000000  0.00000000
#> drat         0.000000  0.00000  0.000000  0.000000  0.000000  0.00000000
#> gear         0.000000  0.00000  0.000000  0.000000  0.000000  0.00000000
#> hp           0.000000  0.00000  0.000000  0.000000  0.000000  0.00000000
#> qsec         0.000000  0.00000  0.000000  0.000000  0.000000  0.04646137
#> vs           0.000000  0.00000  0.000000  0.000000  0.000000  0.00000000
#> wt          -4.849708 -5.09919 -5.331858 -5.548845 -5.700939 -5.68245965
#>                 1.5498     1.4453     1.3479     1.2571     1.1724     1.0933
#> (Intercept) 35.6831165 33.8445365 32.1298706 30.5307675 29.0398995 27.6490503
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.1486318  0.2439042  0.3327556  0.4156186  0.4928731  0.5649448
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.6418236 -5.6039311 -5.5685924 -5.5356354 -5.5049091 -5.4762442
#>                 1.0197     0.9509    0.8868     0.8271     0.7713     0.7193
#> (Intercept) 26.3519394 25.1422495 24.014089 22.9619621 21.9807453 21.0656593
#> am           0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> qsec         0.6321591  0.6948434  0.753303  0.8078226  0.8586678  0.9060861
#> vs           0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> wt          -5.4495111 -5.4245798 -5.401329 -5.3796448 -5.3594223 -5.3405627
#>                 0.6709     0.6257    0.5835     0.5442    0.5075    0.4733
#> (Intercept) 20.2122470 19.4163518 18.939014 16.4996094 14.045339 11.754620
#> am           0.0000000  0.0000000  0.000000  0.7602283  1.524761  2.238322
#> carb         0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.000000
#> cyl          0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.000000
#> disp         0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.000000
#> drat         0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.000000
#> gear         0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.000000
#> hp           0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.000000
#> qsec         0.9503086  0.9915507  1.016286  1.0810335  1.146202  1.207031
#> vs           0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.000000
#> wt          -5.3229741 -5.3065710 -5.296733 -4.9975134 -4.696577 -4.415704
#>                  0.4414     0.4116     0.3839     0.3580    0.3339    0.3114
#> (Intercept) 10.45703522 11.1232500 11.7438334 12.3231254 12.862824 12.958683
#> am           2.78874972  2.9401533  3.0806425  3.2121820  3.334321  3.356165
#> carb        -0.08501418 -0.2938766 -0.4880693 -0.6696073 -0.838463 -0.868593
#> cyl          0.00000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> disp         0.00000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> drat         0.00000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> gear         0.00000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> hp           0.00000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> qsec         1.23114833  1.1880841  1.1480072  1.1105693  1.075719  1.069519
#> vs           0.00000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> wt          -4.14949973 -3.9681647 -3.7996818 -3.6420933 -3.495601 -3.469436
#>                 0.2904     0.2708     0.2526      0.2356      0.2197
#> (Intercept) 12.9575136 12.9566967 21.1048545 21.10735286 21.10905075
#> am           3.3567115  3.3570922  3.0011192  3.00059518  3.00020917
#> carb        -0.8687586 -0.8688733 -0.8065526 -0.80637713 -0.80624761
#> cyl          0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> disp         0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> drat         0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> gear         0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> hp           0.0000000  0.0000000 -0.0220849 -0.02209025 -0.02209373
#> qsec         1.0695380  1.0695515  0.6627144  0.66262130  0.66256037
#> vs           0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> wt          -3.4691179 -3.4688965 -2.7835985 -2.78369823 -2.78379052
#>                  0.2049      0.1911      0.1782      0.1662      0.1550
#> (Intercept) 21.11018438 21.11092452 21.11139370 21.11167903 13.00303772
#> am           2.99992679  2.99972172  2.99957392  2.99946830  3.44815334
#> carb        -0.80615268 -0.80608359 -0.80603369 -0.80599793  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.02633847
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.02209591 -0.02209722 -0.02209795 -0.02209831 -0.04276547
#> qsec         0.66252160  0.66249791  0.66248428  0.66247723  1.22694072
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -2.78387235 -2.78394275 -2.78400197 -2.78405089 -5.15706585
#>                  0.1445      0.1348      0.1257      0.1172      0.1093
#> (Intercept) 12.99973143 12.99687863 12.99442186 12.99231032 12.99049921
#> am           3.44847144  3.44874409  3.44897728  3.44917629  3.44934571
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.02634208  0.02634512  0.02634768  0.02634982  0.02635160
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.04276186 -0.04275865 -0.04275580 -0.04275328 -0.04275105
#> qsec         1.22711660  1.22726799  1.22739805  1.22750954  1.22760491
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -5.15748050 -5.15783481 -5.15813686 -5.15839376 -5.15861169
#>                  0.1020      0.0951      0.0887      0.0827      0.0771
#> (Intercept) 12.98894907 13.02322857 13.02617692 13.02882054 -0.07897056
#> am           3.44948960  3.09045126  3.09003833  3.08966650  2.99988970
#> carb         0.00000000 -0.33786705 -0.33799145 -0.33810200 -0.93386132
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.88343399
#> disp         0.02635309  0.02459958  0.02459573  0.02459231  0.01478659
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  2.73399181
#> gear         0.00000000  0.57522082  0.57542940  0.57561759  0.00000000
#> hp          -0.04274909 -0.04196985 -0.04197275 -0.04197541 -0.04362033
#> qsec         1.22768632  1.10109348  1.10089136  1.10071022  1.06995033
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -5.15879607 -4.70294558 -4.70241056 -4.70193342 -3.30591482
#>                  0.0719      0.0671      0.0626      0.0583      0.0544
#> (Intercept) -0.08401297 -0.08863152 -0.09285076 -0.09669419 -0.10018433
#> am           3.00008060  3.00027043  3.00045873  3.00064510  3.00082916
#> carb        -0.93389127 -0.93390787 -0.93391228 -0.93390557 -0.93388876
#> cyl          0.88371926  0.88397578  0.88420537  0.88440979  0.88459069
#> disp         0.01478900  0.01479149  0.01479404  0.01479665  0.01479930
#> drat         2.73440975  2.73476833  2.73507181  2.73532426  2.73552952
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.04362565 -0.04363062 -0.04363525 -0.04363955 -0.04364354
#> qsec         1.07008015  1.07020666  1.07032981  1.07044953  1.07056578
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -3.30603604 -3.30618059 -3.30634595 -3.30652977 -3.30672986
#>                  0.0507      0.0473      0.0441      0.0412      0.0384
#> (Intercept) -0.10334266 -0.10618977 -0.10874525 -0.11102781 -0.11305529
#> am           3.00101057  3.00118905  3.00136434  3.00153621  3.00170448
#> carb        -0.93386280 -0.93382859 -0.93378696 -0.93373870 -0.93368455
#> cyl          0.88474965  0.88488818  0.88500771  0.88510959  0.88519511
#> disp         0.01480198  0.01480468  0.01480739  0.01481010  0.01481280
#> drat         2.73569123  2.73581281  2.73589752  2.73594841  2.73596837
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.04364724 -0.04365067 -0.04365383 -0.04365674 -0.04365942
#> qsec         1.07067853  1.07078776  1.07089349  1.07099570  1.07109443
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -3.30694419 -3.30717086 -3.30740811 -3.30765429 -3.30790790
#>                  0.0358      0.0334      0.0311      0.0290      0.0271
#> (Intercept) -0.63824512 -0.63421681 -0.63031109 -0.62652808 -0.62286753
#> am           2.95242698  2.95257605  2.95271536  2.95284538  2.95296653
#> carb        -0.92848377 -0.92825489 -0.92803705 -0.92782990 -0.92763310
#> cyl          0.91404104  0.91372488  0.91341989  0.91312593  0.91284288
#> disp         0.01530691  0.01531025  0.01531338  0.01531632  0.01531907
#> drat         2.49316495  2.49239520  2.49166139  2.49096245  2.49029729
#> gear         0.26864388  0.26860870  0.26857191  0.26853373  0.26849438
#> hp          -0.04438376 -0.04438102 -0.04437833 -0.04437571 -0.04437314
#> qsec         1.10035824  1.10039909  1.10043563  1.10046813  1.10049685
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -3.39482416 -3.39535484 -3.39585629 -3.39632963 -3.39677596
#>                  0.0253      0.0236      0.0220      0.0205      0.0191
#> (Intercept) -0.61932889 -0.61591134 -1.17793244 -1.18173475 -1.18543500
#> am           2.95307925  2.95318395  2.92503294  2.92485476  2.92468510
#> carb        -0.92744631 -0.92726918 -0.95548411 -0.95573646 -0.95597916
#> cyl          0.91257056  0.91230880  0.87998686  0.88028184  0.88057002
#> disp         0.01532164  0.01532405  0.01458509  0.01458083  0.01457676
#> drat         2.48966486  2.48906408  2.61421943  2.61505783  2.61586416
#> gear         0.26845405  0.26841294  0.20825392  0.20824244  0.20823531
#> hp          -0.04437063 -0.04436819 -0.04313052 -0.04313211 -0.04313374
#> qsec         1.10052202  1.10054388  1.13303154  1.13297269  1.13291659
#> vs           0.00000000  0.00000000 -0.31503009 -0.31506876 -0.31509817
#> wt          -3.39719634 -3.39759183 -3.35029251 -3.34966977 -3.34907286
#>                  0.0178      0.0166      0.0155      0.0145      0.0135
#> (Intercept) -1.18903314 -1.19252928 -1.19592376 -1.19921706 -1.20240983
#> am           2.92452372  2.92437035  2.92422475  2.92408666  2.92395583
#> carb        -0.95621241 -0.95643640 -0.95665134 -0.95685745 -0.95705493
#> cyl          0.88085131  0.88112564  0.88139297  0.88165328  0.88190653
#> disp         0.01457289  0.01456920  0.01456570  0.01456238  0.01455923
#> drat         2.61663906  2.61738321  2.61809728  2.61878197  2.61943798
#> gear         0.20823225  0.20823298  0.20823725  0.20824481  0.20825542
#> hp          -0.04313540 -0.04313708 -0.04313879 -0.04314051 -0.04314224
#> qsec         1.13286316  1.13281232  1.13276399  1.13271810  1.13267457
#> vs          -0.31511890 -0.31513149 -0.31513648 -0.31513438 -0.31512567
#> wt          -3.34850116 -3.34795402 -3.34743082 -3.34693092 -3.34645367
#>                  0.0126      0.0117      0.0109      0.0102      0.0095
#> (Intercept) -1.20550286 -1.20849709 -1.21139355 -1.21419341 -1.21689795
#> am           2.92383202  2.92371499  2.92360450  2.92350031  2.92340219
#> carb        -0.95724399 -0.95742486 -0.95759774 -0.95776285 -0.95792041
#> cyl          0.88215274  0.88239192  0.88262409  0.88284928  0.88306755
#> disp         0.01455624  0.01455341  0.01455074  0.01454822  0.01454584
#> drat         2.62006602  2.62066680  2.62124103  2.62178945  2.62231276
#> gear         0.20826884  0.20828487  0.20830330  0.20832391  0.20834654
#> hp          -0.04314398 -0.04314572 -0.04314747 -0.04314920 -0.04315094
#> qsec         1.13263331  1.13259425  1.13255732  1.13252244  1.13248954
#> vs          -0.31511082 -0.31509029 -0.31506450 -0.31503385 -0.31499874
#> wt          -3.34599845 -3.34556460 -3.34515151 -3.34475853 -3.34438504
#>                  0.0089      0.0083      0.0077      0.0072      0.0067
#> (Intercept) -1.21950852 -1.22202658 -1.22445364 -1.22679131 -1.22904123
#> am           2.92330992  2.92322326  2.92314201  2.92306595  2.92299487
#> carb        -0.95807064 -0.95821375 -0.95834995 -0.95847946 -0.95860248
#> cyl          0.88327896  0.88348356  0.88368144  0.88387267  0.88405736
#> disp         0.01454360  0.01454149  0.01453951  0.01453765  0.01453591
#> drat         2.62281168  2.62328693  2.62373922  2.62416925  2.62457774
#> gear         0.20837099  0.20839709  0.20842469  0.20845362  0.20848375
#> hp          -0.04315266 -0.04315437 -0.04315606 -0.04315774 -0.04315940
#> qsec         1.13245854  1.13242938  1.13240198  1.13237628  1.13235220
#> vs          -0.31495954 -0.31491660 -0.31487027 -0.31482086 -0.31476867
#> wt          -3.34403043 -3.34369407 -3.34337537 -3.34307371 -3.34278852
#>                  0.0063      0.0058      0.0054
#> (Intercept) -1.23120512 -1.23328474 -1.23528187
#> am           2.92292856  2.92286683  2.92280948
#> carb        -0.95871923 -0.95882991 -0.95893473
#> cyl          0.88423558  0.88440745  0.88457307
#> disp         0.01453429  0.01453277  0.01453136
#> drat         2.62496536  2.62533282  2.62568078
#> gear         0.20851492  0.20854702  0.20857991
#> hp          -0.04316104 -0.04316266 -0.04316425
#> qsec         1.13232969  1.13230869  1.13228912
#> vs          -0.31471400 -0.31465712 -0.31459829
#> wt          -3.34251921 -3.34226521 -3.34202595
#> 
#> $iter
#>   [1]   0   6   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   4
#>  [19]   4   4   4   4   3   3   3   3   3   3   3   3   3   3   3  33  32  32
#>  [37]  34  26  25  25  24   9   1   1  30   1   1   1   1   1   1  80   1   1
#>  [55]   1   1   1   1  62   1   1 171   1   1   1   1   1   1   1   1   1   1
#>  [73]  78   1   1   1   1   1   1  81   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.441598418 5.074852897 4.732824797 4.413848246 4.116369647 3.838940110
#>   [7] 3.580208395 3.338914331 3.113882680 2.904017410 2.708296356 2.525766246
#>  [13] 2.355538054 2.196782673 2.048726873 1.910649539 1.781878155 1.661785531
#>  [19] 1.549786748 1.445336308 1.347925478 1.257079812 1.172356840 1.093343913
#>  [25] 1.019656192 0.950934778 0.886844956 0.827074575 0.771332516 0.719347286
#>  [31] 0.670865687 0.625651585 0.583484762 0.544159842 0.507485290 0.473282480
#>  [37] 0.441384825 0.411636965 0.383894011 0.358020839 0.333891433 0.311388268
#>  [43] 0.290401741 0.270829636 0.252576625 0.235553805 0.219678267 0.204872687
#>  [49] 0.191064954 0.178187815 0.166178553 0.154978675 0.144533631 0.134792548
#>  [55] 0.125707981 0.117235684 0.109334391 0.101965619 0.095093478 0.088684496
#>  [61] 0.082707457 0.077133252 0.071934729 0.067086569 0.062565158 0.058348476
#>  [67] 0.054415984 0.050748529 0.047328248 0.044138482 0.041163696 0.038389401
#>  [73] 0.035802084 0.033389143 0.031138827 0.029040174 0.027082964 0.025257662
#>  [79] 0.023555381 0.021967827 0.020487269 0.019106495 0.017818782 0.016617855
#>  [85] 0.015497867 0.014453363 0.013479255 0.012570798 0.011723568 0.010933439
#>  [91] 0.010196562 0.009509348 0.008868450 0.008270746 0.007713325 0.007193473
#>  [97] 0.006708657 0.006256516 0.005834848 0.005441598
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 33
#> 
#> $loss
#>   [1] 824.16952 704.79680 604.92377 521.74427 452.83560 396.10735 349.75701
#>   [8] 312.23139 282.19320 258.49197 240.13892 226.28505 216.20221 209.26670
#>  [15] 204.94495 202.78126 202.33866 196.56845 184.82909 175.05857 166.96954
#>  [22] 160.31542 154.88519 150.49223 146.98075 144.21510 142.07870 140.47144
#>  [29] 139.30751 138.51338 138.02619 137.79224 137.75398 129.85255 124.48588
#>  [36] 121.81085 119.52320 115.97655 113.65672 112.34011 111.85456 111.84282
#>  [43] 111.84281 111.84281 104.33525 104.33524 104.33524 104.33524 104.33524
#>  [50] 104.33524 104.33524  94.50444  94.50443  94.50442  94.50441  94.50441
#>  [57]  94.50440  94.50440  92.98190  92.98188  92.98187  87.85992  87.85991
#>  [64]  87.85990  87.85989  87.85988  87.85987  87.85986  87.85986  87.85985
#>  [71]  87.85985  87.85984  87.72170  87.72168  87.72165  87.72164  87.72162
#>  [78]  87.72160  87.72159  87.64648  87.64645  87.64643  87.64640  87.64638
#>  [85]  87.64636  87.64634  87.64633  87.64631  87.64630  87.64628  87.64627
#>  [92]  87.64626  87.64625  87.64624  87.64623  87.64623  87.64622  87.64621
#>  [99]  87.64621  87.64620
#> 
#> $linear.predictors
#>     5.4416   5.0749   4.7328   4.4138   4.1164   3.8389   3.5802   3.3389
#> 1  20.5381 20.84915 21.13924 21.40979 21.66209 21.89740 22.11684 22.32150
#> 2  20.5381 20.70219 20.85522 20.99794 21.13104 21.25516 21.37093 21.47889
#> 3  20.5381 21.02205 21.47339 21.89432 22.28687 22.65296 22.99439 23.31280
#> 4  20.5381 20.37656 20.22590 20.08541 19.95438 19.83218 19.71822 19.61194
#> 5  20.5381 20.30163 20.08111 19.87544 19.68364 19.50477 19.33795 19.18237
#> 6  20.5381 20.52064 20.50436 20.48918 20.47502 20.46182 20.44950 20.43802
#> 7  20.5381 20.54369 20.54892 20.55378 20.55833 20.56256 20.56651 20.57019
#> 8  20.5381 20.37656 20.22590 20.08541 19.95438 19.83218 19.71822 19.61194
#> 9  20.5381 20.01346 19.52419 19.06790 18.64235 18.24549 17.87537 17.53020
#> 10 20.5381 19.33339 18.20987 17.16208 16.18491 15.27359 14.42369 13.63107
#> 11 20.5381 19.23311 18.01607 16.88105 15.82254 14.83536 13.91471 13.05612
#> 12 20.5381 21.09121 21.60705 22.08813 22.53678 22.95519 23.34541 23.70932
#> 13 20.5381 21.42837 22.25864 23.03296 23.75509 24.42855 25.05662 25.64236
#> 14 20.5381 21.30158 22.01360 22.67764 23.29692 23.87447 24.41309 24.91541
#> 15 20.5381 20.93848 21.31189 21.66013 21.98489 22.28777 22.57024 22.83367
#> 16 20.5381 20.33045 20.13680 19.95620 19.78777 19.63069 19.48420 19.34759
#> 17 20.5381 20.37944 20.23147 20.09348 19.96479 19.84477 19.73284 19.62846
#> 18 20.5381 20.14314 19.77480 19.43129 19.11093 18.81216 18.53353 18.27368
#> 19 20.5381 21.12579 21.67388 22.18503 22.66173 23.10631 23.52092 23.90758
#> 20 20.5381 20.53217 20.52664 20.52148 20.51667 20.51219 20.50801 20.50411
#> 21 20.5381 20.75694 20.96103 21.15137 21.32888 21.49443 21.64882 21.79280
#>      3.1139   2.9040   2.7083   2.5258    2.3555    2.1968    2.0487    1.9106
#> 1  22.51236 22.69036 22.85636 23.01117 23.155552 23.290201 23.415775 23.532886
#> 2  21.57957 21.67347 21.76104 21.84271 21.918876 21.989907 22.056151 22.117931
#> 3  23.60975 23.88669 24.14496 24.38583 24.610464 24.819958 25.015332 25.197539
#> 4  19.51282 19.42038 19.33417 19.25377 19.178791 19.108865 19.043652 18.982833
#> 5  19.03728 18.90197 18.77577 18.65809 18.548329 18.445971 18.350510 18.261484
#> 6  20.42731 20.41732 20.40801 20.39932 20.391218 20.383663 20.376616 20.370045
#> 7  20.57363 20.57683 20.57982 20.58261 20.585207 20.587630 20.589890 20.591998
#> 8  19.51282 19.42038 19.33417 19.25377 19.178791 19.108865 19.043652 18.982833
#> 9  17.20829 16.90808 16.62810 16.36699 16.123476 15.896376 15.684581 15.487061
#> 10 12.89188 12.20250 11.55958 10.96000 10.400821  9.879332  9.392989  8.939424
#> 11 12.25539 11.50862 10.81219 10.16269  9.556972  8.992073  8.465246  7.973926
#> 12 24.04871 24.36522 24.66041 24.93569 25.192429 25.431861 25.655155 25.863401
#> 13 26.18863 26.69807 27.17319 27.61628 28.029508 28.414887 28.774292 29.109475
#> 14 25.38387 25.82076 26.22821 26.60820 26.962572 27.293065 27.601283 27.888729
#> 15 23.07934 23.30846 23.52214 23.72141 23.907256 24.080575 24.242213 24.392957
#> 16 19.22018 19.10136 18.99054 18.88720 18.790815 18.700930 18.617103 18.538926
#> 17 19.53111 19.44032 19.35565 19.27668 19.203040 19.134361 19.070311 19.010578
#> 18 18.03134 17.80533 17.59455 17.39798 17.214660 17.043693 16.884249 16.735551
#> 19 24.26819 24.60449 24.91813 25.21063 25.483411 25.737812 25.975067 26.196331
#> 20 20.50047 20.49708 20.49391 20.49096 20.488213 20.485646 20.483253 20.481021
#> 21 21.92708 22.05231 22.16910 22.27802 22.379598 22.474330 22.562678 22.645071
#>       1.7819    1.6618    1.5498    1.4453    1.3479    1.2571    1.1724
#> 1  23.614973 23.531525 23.348018 23.176900 23.017315 22.868486 22.729730
#> 2  22.161234 22.108516 21.992586 21.884484 21.783667 21.689645 21.601987
#> 3  25.325255 25.336154 25.360123 25.382473 25.403317 25.422756 25.440880
#> 4  18.940204 18.897926 18.804956 18.718263 18.637412 18.562011 18.491713
#> 5  18.199082 18.104382 17.896133 17.701945 17.520844 17.351948 17.194485
#> 6  20.365438 20.456996 20.658335 20.846080 21.021172 21.184463 21.336702
#> 7  20.593476 20.819032 21.315040 21.777559 22.208907 22.611183 22.986231
#> 8  18.940204 18.985273 19.084384 19.176803 19.262993 19.343374 19.418315
#> 9  15.348612 15.335632 15.307087 15.280470 15.255646 15.232496 15.210912
#> 10  8.621505  8.657277  8.735942  8.809295  8.877705  8.941505  9.000986
#> 11  7.629541  7.661095  7.730483  7.795187  7.855529  7.911805  7.964272
#> 12 26.009368 26.058006 26.164965 26.264702 26.357718 26.444465 26.525340
#> 13 29.344417 29.338107 29.324232 29.311293 29.299227 29.287974 29.277482
#> 14 28.090210 28.152082 28.288142 28.415016 28.533339 28.643687 28.746567
#> 15 24.498619 24.577244 24.750143 24.911369 25.061729 25.201955 25.332690
#> 16 18.484128 18.436360 18.331315 18.233363 18.142012 18.056817 17.977389
#> 17 18.968708 18.939347 18.874782 18.814576 18.758427 18.706062 18.657242
#> 18 16.631323 16.597924 16.524476 16.455988 16.392115 16.332547 16.277011
#> 19 26.351424 26.270256 26.091764 25.925324 25.770101 25.625339 25.490376
#> 20 20.479457 20.315107 19.953696 19.616686 19.302388 19.009274 18.735998
#> 21 22.702823 22.721758 22.763398 22.802226 22.838437 22.872208 22.903693
#>       1.0933    1.0197    0.9509    0.8868    0.8271    0.7713    0.7193
#> 1  22.600283 22.479560 22.366973 22.261975 22.164053 22.072730 21.987563
#> 2  21.520209 21.443944 21.372818 21.306486 21.244624 21.186932 21.133127
#> 3  25.457787 25.473555 25.488261 25.501975 25.514765 25.526693 25.537817
#> 4  18.426131 18.364970 18.307930 18.254735 18.205125 18.158858 18.115710
#> 5  17.047585 16.910585 16.782819 16.663664 16.552540 16.448905 16.352255
#> 6  21.478728 21.611182 21.734708 21.849910 21.957347 22.057543 22.150987
#> 7  23.336118 23.662424 23.966738 24.250542 24.515218 24.762057 24.992259
#> 8  19.488228 19.553429 19.614236 19.670944 19.723831 19.773153 19.819151
#> 9  15.190777 15.171998 15.154485 15.138153 15.122921 15.108716 15.095468
#> 10  9.056477  9.108227  9.156490  9.201500  9.243477  9.282625  9.319134
#> 11  8.013219  8.058867  8.101438  8.141141  8.178167  8.212698  8.244902
#> 12 26.600789 26.671153 26.736775 26.797975 26.855050 26.908278 26.957918
#> 13 29.267694 29.258566 29.250053 29.242114 29.234710 29.227805 29.221366
#> 14 28.842544 28.932053 29.015530 29.093380 29.165984 29.233694 29.296841
#> 15 25.454655 25.568399 25.674477 25.773406 25.865668 25.951711 26.031956
#> 16 17.903290 17.834185 17.769737 17.709633 17.653580 17.601304 17.552552
#> 17 18.611697 18.569222 18.529609 18.492666 18.458213 18.426082 18.396116
#> 18 16.225201 16.176882 16.131821 16.089796 16.050603 16.014052 15.979964
#> 19 25.364467 25.247043 25.137534 25.035405 24.940160 24.851333 24.768493
#> 20 18.481056 18.243297 18.021561 17.814770 17.621916 17.442059 17.274324
#> 21 22.933065 22.960459 22.986005 23.009830 23.032050 23.052772 23.072097
#>       0.6709    0.6257    0.5835    0.5442    0.5075    0.4733    0.4414
#> 1  21.908135 21.834061 21.789635 21.960164 22.131560 22.291522 22.298741
#> 2  21.082950 21.036154 21.008088 21.291176 21.575806 21.841455 21.930061
#> 3  25.548191 25.557866 25.563669 25.783640 26.004869 26.211349 26.445602
#> 4  18.075469 18.037941 18.015433 17.707353 17.397478 17.108260 16.966872
#> 5  16.262118 16.178057 16.127641 15.782057 15.434404 15.109922 14.804654
#> 6  22.238133 22.319405 22.368148 22.178211 21.987306 21.809138 21.673069
#> 7  25.206947 25.407165 25.527246 25.513109 25.499157 25.486155 25.409380
#> 8  19.862049 19.902056 19.926050 19.739696 19.552339 19.377478 19.111403
#> 9  15.083113 15.071591 15.064680 14.969712 14.874192 14.785038 14.735510
#> 10  9.353182  9.384936  9.403981  9.699646  9.997028 10.274586 10.468152
#> 11  8.274936  8.302945  8.319743  8.657113  8.996431  9.313128  9.549155
#> 12 27.004213 27.047388 27.073282 27.313030 27.554192 27.779280 28.002329
#> 13 29.215360 29.209759 29.206400 29.209593 29.212798 29.215788 29.175182
#> 14 29.355732 29.410653 29.443593 29.601967 29.761310 29.910035 30.046291
#> 15 26.106792 26.176584 26.218442 25.812219 25.403787 25.022593 24.778782
#> 16 17.507085 17.464683 17.439252 17.145397 16.849822 16.573949 16.450240
#> 17 18.368171 18.342108 18.326477 18.035030 17.741898 17.468307 17.332341
#> 18 15.948174 15.918526 15.900745 15.715791 15.529751 15.356111 15.323260
#> 19 24.691237 24.619187 24.575975 24.618418 24.661006 24.700747 24.756004
#> 20 17.117894 16.972007 16.884512 17.092706 17.301886 17.497105 17.603465
#> 21 23.090120 23.106928 23.117009 23.473973 23.832981 24.168055 24.439506
#>       0.4116   0.3839    0.3580    0.3339    0.3114    0.2904    0.2708
#> 1  22.047170 21.81323 21.594564 21.391144 21.354832 21.354697 21.354605
#> 2  21.700615 21.48720 21.287749 21.102168 21.069056 21.069014 21.068985
#> 3  26.673630 26.88556 27.083738 27.268008 27.300907 27.301215 27.301429
#> 4  17.106202 17.23587 17.356999 17.469759 17.489846 17.489768 17.489713
#> 5  14.600648 14.41113 14.233840 14.069057 14.039601 14.039210 14.038939
#> 6  21.638734 21.60685 21.577018 21.549300 21.544371 21.544271 21.544201
#> 7  25.242904 25.08806 24.943353 24.808708 24.784753 24.784695 24.784656
#> 8  18.752047 18.41799 18.105654 17.815184 17.763355 17.762982 17.762723
#> 9  14.763854 14.79025 14.814889 14.837840 14.841925 14.841889 14.841864
#> 10 10.476631 10.48440 10.491741 10.498484 10.499718 10.499904 10.500033
#> 11  9.596077  9.63957  9.680326  9.718134  9.724913  9.725151  9.725317
#> 12 28.171562 28.32881 28.475878 28.612599 28.637026 28.637312 28.637511
#> 13 29.070382 28.97294 28.881855 28.797130 28.782010 28.781926 28.781869
#> 14 30.130819 30.20933 30.282787 30.351052 30.363263 30.363442 30.363566
#> 15 24.821411 24.86117 24.898249 24.932832 24.939000 24.938835 24.938719
#> 16 16.610536 16.75970 16.899046 17.028753 17.051863 17.051808 17.051769
#> 17 17.458706 17.57631 17.686168 17.788438 17.806658 17.806584 17.806532
#> 18 15.534738 15.73144 15.915268 16.086312 16.116810 16.116861 16.116896
#> 19 24.824782 24.88874 24.948520 25.004131 25.014032 25.014080 25.014114
#> 20 17.536034 17.47331 17.414697 17.360155 17.350385 17.350388 17.350391
#> 21 24.542517 24.63816 24.727662 24.810812 24.825678 24.825967 24.826168
#>       0.2526    0.2356    0.2197    0.2049    0.1911    0.1782    0.1662
#> 1  22.065675 22.065969 22.066172 22.066310 22.066403 22.066464 22.066503
#> 2  21.726978 21.727194 21.727339 21.727435 21.727497 21.727535 21.727557
#> 3  27.120692 27.120380 27.120150 27.119982 27.119860 27.119772 27.119710
#> 4  17.330712 17.330697 17.330691 17.330692 17.330695 17.330700 17.330706
#> 5  13.027793 13.027851 13.027921 13.027994 13.028063 13.028125 13.028179
#> 6  22.497094 22.497432 22.497660 22.497812 22.497911 22.497973 22.498010
#> 7  23.801508 23.801403 23.801344 23.801314 23.801304 23.801305 23.801312
#> 8  18.111925 18.112364 18.112683 18.112914 18.113080 18.113197 18.113280
#> 9  14.911899 14.911935 14.911960 14.911978 14.911991 14.912000 14.912006
#> 10 10.652952 10.652858 10.652781 10.652721 10.652673 10.652636 10.652607
#> 11  9.841723  9.841572  9.841455  9.841365  9.841295  9.841242  9.841203
#> 12 28.620950 28.620715 28.620538 28.620405 28.620306 28.620234 28.620180
#> 13 29.122413 29.122574 29.122687 29.122765 29.122817 29.122852 29.122875
#> 14 29.944016 29.943782 29.943616 29.943499 29.943417 29.943361 29.943323
#> 15 24.555411 24.555457 24.555501 24.555540 24.555575 24.555604 24.555629
#> 16 17.560739 17.560864 17.560947 17.561001 17.561036 17.561056 17.561068
#> 17 18.082312 18.082406 18.082470 18.082515 18.082545 18.082565 18.082578
#> 18 16.223236 16.223178 16.223133 16.223099 16.223073 16.223054 16.223040
#> 19 25.593572 25.593642 25.593682 25.593702 25.593710 25.593710 25.593706
#> 20 15.834701 15.834298 15.834035 15.833868 15.833767 15.833709 15.833680
#> 21 24.673698 24.673431 24.673234 24.673088 24.672982 24.672906 24.672850
#>       0.1550   0.1445    0.1348    0.1257    0.1172    0.1093    0.1020
#> 1  22.645076 22.64487 22.644694 22.644542 22.644412 22.644300 22.644204
#> 2  22.017111 22.01690 22.016716 22.016560 22.016427 22.016313 22.016216
#> 3  26.187531 26.18758 26.187622 26.187659 26.187691 26.187719 26.187743
#> 4  18.143153 18.14335 18.143507 18.143644 18.143758 18.143854 18.143934
#> 5  13.031361 13.03154 13.031707 13.031850 13.031976 13.032087 13.032184
#> 6  22.303206 22.30285 22.302538 22.302271 22.302040 22.301841 22.301671
#> 7  24.500959 24.50123 24.501457 24.501658 24.501831 24.501981 24.502110
#> 8  17.606085 17.60573 17.605420 17.605161 17.604942 17.604757 17.604602
#> 9  12.928913 12.92862 12.928381 12.928175 12.928003 12.927858 12.927737
#> 10 11.653671 11.65379 11.653897 11.653984 11.654056 11.654116 11.654165
#> 11  9.816315  9.81633  9.816343  9.816354  9.816363  9.816371  9.816377
#> 12 28.244498 28.24454 28.244584 28.244617 28.244645 28.244669 28.244688
#> 13 30.615489 30.61555 30.615598 30.615637 30.615668 30.615692 30.615710
#> 14 30.497005 30.49725 30.497455 30.497632 30.497783 30.497911 30.498020
#> 15 23.856953 23.85693 23.856908 23.856892 23.856880 23.856871 23.856865
#> 16 17.509468 17.50936 17.509262 17.509176 17.509100 17.509033 17.508974
#> 17 18.106664 18.10662 18.106571 18.106531 18.106495 18.106463 18.106434
#> 18 17.144888 17.14506 17.145207 17.145327 17.145427 17.145509 17.145575
#> 19 25.181840 25.18166 25.181512 25.181381 25.181268 25.181170 25.181086
#> 20 15.848650 15.84912 15.849526 15.849879 15.850186 15.850452 15.850682
#> 21 23.461164 23.46112 23.461093 23.461069 23.461050 23.461035 23.461024
#>      0.0951    0.0887    0.0827   0.0771   0.0719   0.0671   0.0626   0.0583
#> 1  22.18463 22.184637 22.184649 21.66615 21.66614 21.66614 21.66614 21.66616
#> 2  21.60199 21.602021 21.602053 21.42231 21.42234 21.42238 21.42241 21.42244
#> 3  26.41077 26.410810 26.410843 25.82898 25.82874 25.82853 25.82834 25.82816
#> 4  18.14676 18.146591 18.146441 18.26067 18.26090 18.26111 18.26131 18.26151
#> 5  12.62246 12.622151 12.621869 11.81126 11.81091 11.81059 11.81031 11.81007
#> 6  22.67448 22.674932 22.675340 21.99334 21.99316 21.99298 21.99281 21.99266
#> 7  24.52562 24.525398 24.525194 24.33054 24.33064 24.33074 24.33084 24.33094
#> 8  17.56578 17.566079 17.566351 18.16608 18.16605 18.16601 18.16596 18.16590
#> 9  12.98332 12.983596 12.983841 13.96882 13.96872 13.96861 13.96849 13.96836
#> 10 11.51180 11.511639 11.511496 11.18241 11.18250 11.18259 11.18268 11.18277
#> 11  9.80242  9.802401  9.802384 10.01372 10.01372 10.01371 10.01370 10.01368
#> 12 28.33448 28.334475 28.334468 28.51916 28.51923 28.51928 28.51934 28.51938
#> 13 30.21558 30.215378 30.215201 31.39303 31.39343 31.39380 31.39414 31.39446
#> 14 30.37954 30.379284 30.379053 30.49990 30.50011 30.50031 30.50050 30.50069
#> 15 23.73448 23.734458 23.734439 23.44200 23.44168 23.44139 23.44113 23.44088
#> 16 17.62142 17.621563 17.621688 17.23892 17.23899 17.23905 17.23911 17.23917
#> 17 18.15025 18.150311 18.150365 18.83925 18.83951 18.83974 18.83995 18.84014
#> 18 17.25908 17.258970 17.258872 17.35396 17.35421 17.35444 17.35466 17.35487
#> 19 25.77808 25.778452 25.778787 25.30141 25.30120 25.30100 25.30082 25.30064
#> 20 16.25025 16.249974 16.249729 16.49925 16.49926 16.49927 16.49927 16.49928
#> 21 23.54681 23.546881 23.546939 23.56883 23.56857 23.56832 23.56808 23.56785
#>      0.0544   0.0507   0.0473   0.0441   0.0412   0.0384    0.0358    0.0334
#> 1  21.66618 21.66621 21.66624 21.66628 21.66632 21.66637 21.666762 21.666932
#> 2  21.42248 21.42251 21.42255 21.42259 21.42262 21.42266 21.417283 21.417340
#> 3  25.82800 25.82785 25.82772 25.82760 25.82750 25.82740 25.842255 25.842436
#> 4  18.26169 18.26186 18.26203 18.26218 18.26233 18.26247 18.269751 18.269726
#> 5  11.80985 11.80967 11.80951 11.80938 11.80927 11.80918 11.715761 11.716222
#> 6  21.99251 21.99237 21.99224 21.99212 21.99200 21.99189 22.106713 22.106736
#> 7  24.33104 24.33113 24.33123 24.33132 24.33141 24.33150 24.451998 24.452054
#> 8  18.16583 18.16575 18.16566 18.16557 18.16547 18.16537 18.204661 18.204391
#> 9  13.96821 13.96806 13.96791 13.96775 13.96758 13.96741 13.910450 13.910130
#> 10 11.18287 11.18297 11.18307 11.18317 11.18327 11.18338 11.158861 11.158998
#> 11 10.01367 10.01366 10.01364 10.01363 10.01361 10.01359  9.939106  9.939077
#> 12 28.51942 28.51946 28.51949 28.51951 28.51953 28.51955 28.519239 28.519170
#> 13 31.39474 31.39500 31.39524 31.39545 31.39565 31.39582 31.226029 31.225758
#> 14 30.50086 30.50103 30.50119 30.50134 30.50149 30.50163 30.508598 30.508605
#> 15 23.44065 23.44045 23.44026 23.44008 23.43993 23.43978 23.303140 23.303354
#> 16 17.23923 17.23928 17.23934 17.23939 17.23944 17.23949 17.327481 17.327499
#> 17 18.84031 18.84046 18.84059 18.84070 18.84080 18.84088 18.847232 18.846966
#> 18 17.35507 17.35525 17.35543 17.35559 17.35575 17.35589 17.365613 17.365562
#> 19 25.30047 25.30030 25.30015 25.30000 25.29986 25.29973 25.414878 25.414859
#> 20 16.49928 16.49928 16.49928 16.49927 16.49927 16.49926 16.625968 16.625910
#> 21 23.56764 23.56744 23.56725 23.56707 23.56690 23.56673 23.478221 23.478274
#>      0.0311    0.0290    0.0271    0.0253    0.0236    0.0220    0.0205
#> 1  21.66709 21.667246 21.667391 21.667529 21.667659 21.694748 21.694556
#> 2  21.41739 21.417444 21.417492 21.417536 21.417578 21.474921 21.474855
#> 3  25.84261 25.842780 25.842943 25.843100 25.843251 25.771385 25.771215
#> 4  18.26970 18.269671 18.269641 18.269610 18.269577 18.272531 18.272516
#> 5  11.71667 11.717094 11.717506 11.717901 11.718281 11.726765 11.726334
#> 6  22.10676 22.106783 22.106807 22.106830 22.106854 22.034241 22.034208
#> 7  24.45211 24.452150 24.452191 24.452226 24.452257 24.545956 24.545892
#> 8  18.20414 18.203892 18.203661 18.203442 18.203235 18.074476 18.074755
#> 9  13.90983 13.909542 13.909273 13.909019 13.908781 13.984060 13.984446
#> 10 11.15913 11.159249 11.159363 11.159471 11.159573 11.149730 11.149571
#> 11  9.93905  9.939027  9.939005  9.938986  9.938969  9.962164  9.962216
#> 12 28.51910 28.519040 28.518978 28.518918 28.518861 28.486278 28.486344
#> 13 31.22550 31.225247 31.225006 31.224775 31.224554 31.196494 31.196747
#> 14 30.50861 30.508608 30.508605 30.508599 30.508591 30.494613 30.494578
#> 15 23.30356 23.303769 23.303969 23.304164 23.304354 23.350374 23.350219
#> 16 17.32751 17.327527 17.327536 17.327543 17.327548 17.280696 17.280632
#> 17 18.84671 18.846463 18.846225 18.845996 18.845775 18.868029 18.868273
#> 18 17.36551 17.365457 17.365404 17.365350 17.365296 17.350062 17.350068
#> 19 25.41484 25.414829 25.414818 25.414809 25.414802 25.560052 25.560095
#> 20 16.62585 16.625798 16.625745 16.625692 16.625641 16.579776 16.579829
#> 21 23.47833 23.478385 23.478443 23.478502 23.478563 23.442651 23.442653
#>       0.0191   0.0178    0.0166    0.0155    0.0145    0.0135    0.0126
#> 1  21.694371 21.69419 21.694020 21.693854 21.693694 21.693540 21.693391
#> 2  21.474791 21.47473 21.474667 21.474607 21.474549 21.474492 21.474436
#> 3  25.771052 25.77089 25.770742 25.770595 25.770454 25.770319 25.770188
#> 4  18.272503 18.27249 18.272486 18.272481 18.272479 18.272478 18.272480
#> 5  11.725914 11.72551 11.725108 11.724722 11.724347 11.723983 11.723630
#> 6  22.034176 22.03415 22.034121 22.034096 22.034073 22.034052 22.034033
#> 7  24.545830 24.54577 24.545711 24.545654 24.545598 24.545544 24.545492
#> 8  18.075027 18.07529 18.075550 18.075801 18.076045 18.076282 18.076512
#> 9  13.984816 13.98517 13.985505 13.985825 13.986130 13.986421 13.986697
#> 10 11.149418 11.14927 11.149130 11.148996 11.148867 11.148744 11.148627
#> 11  9.962264  9.96231  9.962352  9.962391  9.962428  9.962461  9.962492
#> 12 28.486410 28.48647 28.486537 28.486599 28.486660 28.486719 28.486777
#> 13 31.196994 31.19723 31.197469 31.197697 31.197919 31.198135 31.198344
#> 14 30.494547 30.49452 30.494494 30.494473 30.494454 30.494438 30.494425
#> 15 23.350063 23.34991 23.349754 23.349601 23.349449 23.349298 23.349149
#> 16 17.280573 17.28052 17.280469 17.280425 17.280384 17.280348 17.280316
#> 17 18.868510 18.86874 18.868965 18.869183 18.869394 18.869599 18.869797
#> 18 17.350076 17.35009 17.350100 17.350116 17.350134 17.350153 17.350174
#> 19 25.560131 25.56016 25.560188 25.560208 25.560224 25.560235 25.560242
#> 20 16.579882 16.57993 16.579987 16.580040 16.580092 16.580144 16.580195
#> 21 23.442653 23.44265 23.442645 23.442637 23.442628 23.442616 23.442603
#>       0.0117    0.0109    0.0102    0.0095    0.0089   0.0083    0.0077
#> 1  21.693249 21.693112 21.692980 21.692854 21.692733 21.69262 21.692506
#> 2  21.474383 21.474330 21.474279 21.474230 21.474182 21.47414 21.474091
#> 3  25.770063 25.769942 25.769827 25.769716 25.769610 25.76951 25.769412
#> 4  18.272483 18.272488 18.272494 18.272502 18.272511 18.27252 18.272533
#> 5  11.723288 11.722957 11.722636 11.722327 11.722027 11.72174 11.721460
#> 6  22.034015 22.033999 22.033985 22.033972 22.033960 22.03395 22.033941
#> 7  24.545441 24.545391 24.545344 24.545298 24.545253 24.54521 24.545168
#> 8  18.076735 18.076952 18.077161 18.077364 18.077560 18.07775 18.077933
#> 9  13.986959 13.987207 13.987443 13.987666 13.987876 13.98808 13.988263
#> 10 11.148515 11.148409 11.148308 11.148211 11.148119 11.14803 11.147950
#> 11  9.962521  9.962547  9.962571  9.962592  9.962612  9.96263  9.962645
#> 12 28.486834 28.486890 28.486944 28.486997 28.487049 28.48710 28.487148
#> 13 31.198547 31.198744 31.198934 31.199119 31.199297 31.19947 31.199636
#> 14 30.494414 30.494406 30.494399 30.494395 30.494392 30.49439 30.494393
#> 15 23.349001 23.348856 23.348712 23.348571 23.348432 23.34830 23.348161
#> 16 17.280288 17.280263 17.280242 17.280224 17.280209 17.28020 17.280187
#> 17 18.869989 18.870175 18.870354 18.870527 18.870695 18.87086 18.871011
#> 18 17.350197 17.350221 17.350246 17.350272 17.350300 17.35033 17.350357
#> 19 25.560245 25.560244 25.560240 25.560232 25.560222 25.56021 25.560193
#> 20 16.580246 16.580297 16.580346 16.580395 16.580444 16.58049 16.580538
#> 21 23.442588 23.442572 23.442555 23.442536 23.442516 23.44250 23.442474
#>       0.0072    0.0067    0.0063    0.0058   0.0054
#> 1  21.692400 21.692299 21.692202 21.692110 21.69202
#> 2  21.474047 21.474005 21.473965 21.473925 21.47389
#> 3  25.769319 25.769230 25.769146 25.769066 25.76899
#> 4  18.272545 18.272559 18.272573 18.272588 18.27260
#> 5  11.721191 11.720932 11.720683 11.720443 11.72021
#> 6  22.033932 22.033926 22.033920 22.033915 22.03391
#> 7  24.545128 24.545089 24.545052 24.545016 24.54498
#> 8  18.078110 18.078281 18.078445 18.078603 18.07876
#> 9  13.988440 13.988606 13.988763 13.988909 13.98905
#> 10 11.147872 11.147798 11.147728 11.147662 11.14760
#> 11  9.962659  9.962672  9.962683  9.962692  9.96270
#> 12 28.487196 28.487242 28.487287 28.487331 28.48737
#> 13 31.199797 31.199952 31.200101 31.200245 31.20038
#> 14 30.494395 30.494399 30.494404 30.494411 30.49442
#> 15 23.348030 23.347901 23.347775 23.347651 23.34753
#> 16 17.280179 17.280174 17.280172 17.280171 17.28017
#> 17 18.871161 18.871305 18.871444 18.871577 18.87171
#> 18 17.350386 17.350416 17.350447 17.350478 17.35051
#> 19 25.560175 25.560155 25.560133 25.560109 25.56008
#> 20 16.580584 16.580630 16.580674 16.580717 16.58076
#> 21 23.442451 23.442428 23.442405 23.442381 23.44236
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 22.8 18.7 14.3 24.4 22.8 17.8 16.4 10.4 10.4 32.4 30.4 33.9 21.5
#> [16] 15.5 15.2 19.2 26.0 15.8 21.4
#> 
#> $X
#>               am       carb       cyl       disp        drat       gear
#>  [1,]  1.1547005  1.2961751  0.000000 -0.5164336  0.39512458  0.5188745
#>  [2,]  1.1547005  1.2961751  0.000000 -0.5164336  0.39512458  0.5188745
#>  [3,]  1.1547005 -1.3379872 -1.080123 -0.9161681  0.30647483  0.5188745
#>  [4,] -0.8660254 -0.4599331  1.080123  1.0210069 -0.93462159 -1.0377490
#>  [5,] -0.8660254  1.2961751  1.080123  1.0210069 -0.82824190 -1.0377490
#>  [6,] -0.8660254 -0.4599331 -1.080123 -0.6186734  0.02279565  0.5188745
#>  [7,] -0.8660254 -0.4599331 -1.080123 -0.6640279  0.43058447  0.5188745
#>  [8,] -0.8660254  1.2961751  0.000000 -0.4580109  0.43058447  0.5188745
#>  [9,] -0.8660254  0.4181210  1.080123  0.3737445 -1.07646119 -1.0377490
#> [10,] -0.8660254  1.2961751  1.080123  1.8819736 -1.32468047 -1.0377490
#> [11,] -0.8660254  1.2961751  1.080123  1.7897272 -1.20057083 -1.0377490
#> [12,]  1.1547005 -1.3379872 -1.080123 -1.1414032  0.71426366  0.5188745
#> [13,]  1.1547005 -0.4599331 -1.080123 -1.1644648  2.22130932  0.5188745
#> [14,]  1.1547005 -1.3379872 -1.080123 -1.1998259  0.96248294  0.5188745
#> [15,] -0.8660254 -1.3379872 -1.080123 -0.8231530  0.04052560 -1.0377490
#> [16,] -0.8660254 -0.4599331  1.080123  0.6981444 -1.62608960 -1.0377490
#> [17,] -0.8660254 -0.4599331  1.080123  0.5905236 -0.93462159 -1.0377490
#> [18,] -0.8660254 -0.4599331  1.080123  1.3284950 -1.05873124 -1.0377490
#> [19,]  1.1547005 -0.4599331 -1.080123 -0.8216155  1.33481187  2.0754981
#> [20,]  1.1547005  1.2961751  1.080123  0.9518221  0.96248294  2.0754981
#> [21,]  1.1547005 -0.4599331 -1.080123 -0.8162345  0.76745350  0.5188745
#>               hp        qsec         vs          wt
#>  [1,] -0.4120879 -0.89505045 -0.8660254 -0.56543644
#>  [2,] -0.4120879 -0.57810395 -0.8660254 -0.29828344
#>  [3,] -0.6939181  0.32179772  1.1547005 -0.87973409
#>  [4,]  0.6654982 -0.57810395 -0.8660254  0.29364380
#>  [5,]  1.8259755 -1.24595551 -0.8660254  0.42983945
#>  [6,] -1.2078437  1.10850422  1.1547005  0.03172910
#>  [7,] -0.6607616  2.74983431  1.1547005 -0.01017726
#>  [8,] -0.1965706  0.48593073  1.1547005  0.29364380
#>  [9,]  0.7483894 -0.36303311 -0.8660254  0.95366887
#> [10,]  1.1628456 -0.03476709 -0.8660254  2.18990629
#> [11,]  1.3286281 -0.12532323 -0.8660254  2.37219893
#> [12,] -1.1415307  0.80853699  1.1547005 -1.00545315
#> [13,] -1.3736262  0.27085989  1.1547005 -1.61833357
#> [14,] -1.1581090  1.05190663  1.1547005 -1.38784862
#> [15,] -0.6276051  1.11416398  1.1547005 -0.72782356
#> [16,]  0.2510420 -0.66300033 -0.8660254  0.37745651
#> [17,]  0.2510420 -0.41963070 -0.8660254  0.28840551
#> [18,]  0.6654982 -0.56112467 -0.8660254  0.71794563
#> [19,] -0.7270745 -0.75921624 -0.8660254 -1.06831268
#> [20,]  2.1409622 -2.00436321 -0.8660254  0.01077592
#> [21,] -0.4286661  0.31613796  1.1547005 -0.39781103
#> attr(,"center")
#>  [1]   0.4285714   2.5238095   6.0000000 227.1809524   3.6771429   3.6666667
#>  [7] 134.8571429  18.0414286   0.4285714   3.1597143
#> attr(,"scale")
#>  [1]   0.4948717   1.1388820   1.8516402 130.0863326   0.5640174   0.6424161
#>  [7]  60.3200083   1.7668597   0.4948717   0.9545092
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 33
#> 
#> $lambda.min
#> [1] 0.5834848
#> 
#> $null.dev
#> [1] 39.24617
#> 
#> $Bias
#> [1] 6.123659
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 5.527033