Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 38.328261 35.964820 31.387888 26.568736 22.580572 19.301459 16.626264
#>   [8] 14.464410 12.737921 11.379725 10.332186  9.545826  8.978220  8.593035
#>  [15]  8.359198  8.239011  8.112262  8.067586  7.982617  7.809807  7.568088
#>  [22]  7.280477  7.053531  6.840688  6.532320  6.278081  6.069978  5.901125
#>  [29]  5.765588  5.658269  5.574778  5.511347  5.456743  5.380366  5.305784
#>  [36]  5.234363  5.177700  5.130380  5.087253  5.068247  5.068246  5.068246
#>  [43]  7.050429  8.180641  8.679857  8.882203 12.393951 12.881239 12.993961
#>  [50] 12.921547 13.314988 13.883973 12.949082 13.651929 13.509349 13.317436
#>  [57] 13.680087 13.373857 13.515968 14.005427 14.005866 13.941404 13.787937
#>  [64] 16.822816 16.890559 19.415602 19.458861 19.353271 19.327279 19.349103
#>  [71] 19.186888 19.188497 18.758987 18.759633 18.760264 18.702059 18.702808
#>  [78] 18.703546 18.704270 18.704983 18.705683 18.706372 18.707048 18.707714
#>  [85] 18.164746 18.164779 18.164823 18.164877 18.164941 18.166520 18.173765
#>  [92] 18.173795 18.173831 18.173873 18.173920 18.173973 18.174031 18.174095
#>  [99] 18.174165 18.174240
#> 
#> $cvse
#>   [1] 10.758029 10.571366  9.608549  8.330268  7.300366  6.469375  5.794247
#>   [8]  5.238724  4.773510  4.376045  4.029886  3.723825  3.450922  3.207566
#>  [15]  2.992641  2.807078  2.656761  2.530457  2.458322  2.415616  2.374765
#>  [22]  2.322805  2.287091  2.240087  2.113072  2.004005  1.910873  1.831883
#>  [29]  1.765422  1.710027  1.664367  1.627226  1.597394  1.570019  1.540051
#>  [36]  1.509784  1.481584  1.455631  1.432219  1.421977  1.421976  1.421976
#>  [43]  2.373300  2.801560  2.807643  2.803324  3.129591  3.325634  3.316194
#>  [50]  3.313730  3.227323  3.153443  2.841515  2.946899  2.889572  2.906443
#>  [57]  2.887178  2.908957  2.907429  2.999044  2.999168  2.938966  2.977400
#>  [64]  4.445266  4.560082  6.199667  6.199613  6.182890  6.185900  6.187556
#>  [71]  6.193102  6.193854  5.959512  5.959972  5.960428  5.938913  5.939357
#>  [78]  5.939799  5.940241  5.940680  5.941118  5.941554  5.941988  5.942420
#>  [85]  5.858827  5.859183  5.859539  5.859896  5.860253  5.860681  5.861458
#>  [92]  5.861823  5.862189  5.862554  5.862917  5.863278  5.863638  5.863996
#>  [99]  5.864351  5.864704
#> 
#> $fold
#>  [1]  1 10 10  6  6  8  9  3  5  7  7  2  4  3  4  2  9  1  8  5  1
#> 
#> $lambda
#>   [1] 5.405683187 5.041358233 4.701587560 4.384716295 4.089201091 3.813602624
#>   [7] 3.556578571 3.316877079 3.093330665 2.884850531 2.690421260 2.509095872
#>  [13] 2.339991208 2.182283632 2.035205019 1.898039012 1.770117536 1.650817538
#>  [19] 1.539557961 1.435796907 1.339029001 1.248782928 1.164619138 1.086127706
#>  [25] 1.012926334 0.944658489 0.880991669 0.821615779 0.766241625 0.714599504
#>  [31] 0.666437889 0.621522206 0.579633689 0.540568319 0.504135823 0.470158756
#>  [37] 0.438471629 0.408920109 0.381360262 0.355657857 0.331687708 0.309333066
#>  [43] 0.288485053 0.269042126 0.250909587 0.233999121 0.218228363 0.203520502
#>  [49] 0.189803901 0.177011754 0.165081754 0.153955796 0.143579691 0.133902900
#>  [55] 0.124878293 0.116461914 0.108612771 0.101292633 0.094465849 0.088099167
#>  [61] 0.082161578 0.076624163 0.071459950 0.066643789 0.062152221 0.057963369
#>  [67] 0.054056832 0.050413582 0.047015876 0.043847163 0.040892011 0.038136026
#>  [73] 0.035565786 0.033168771 0.030933307 0.028848505 0.026904213 0.025090959
#>  [79] 0.023399912 0.021822836 0.020352050 0.018980390 0.017701175 0.016508175
#>  [85] 0.015395580 0.014357969 0.013390290 0.012487829 0.011646191 0.010861277
#>  [91] 0.010129263 0.009446585 0.008809917 0.008216158 0.007662416 0.007145995
#>  [97] 0.006664379 0.006215222 0.005796337 0.005405683
#> 
#> $fit
#> $beta
#>               5.4057     5.0414    4.7016    4.3847    4.0892   3.8136
#> (Intercept) 20.70476 22.5328212 24.237675 25.827628 27.310423 28.69328
#> am           0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> carb         0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> cyl          0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> disp         0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> drat         0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> gear         0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> hp           0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> qsec         0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> vs           0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> wt           0.00000 -0.5972842 -1.154313 -1.673801 -2.158276 -2.61010
#>                3.5566    3.3169    3.0933    2.8849    2.6904    2.5091
#> (Intercept) 29.982943 31.185684 32.307364 33.353447 34.329027 35.238856
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -3.031472 -3.424445 -3.790933 -4.132721 -4.451474 -4.748743
#>                2.3400    2.1823    2.0352    1.8980    1.7701    1.6508
#> (Intercept) 36.087366 36.878690 37.616680 38.304933 38.787354 38.787354
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -5.025978 -5.284528 -5.525653 -5.750526 -5.908149 -5.908149
#>                  1.5396     1.4358    1.3390    1.2488     1.1646     1.0861
#> (Intercept) 37.85363091 35.9165853 34.110087 32.425340 30.8541403 29.3888338
#> am           0.00000000  0.0000000  0.000000  0.000000  0.0000000  0.0000000
#> carb         0.00000000  0.0000000  0.000000  0.000000  0.0000000  0.0000000
#> cyl          0.00000000  0.0000000  0.000000  0.000000  0.0000000  0.0000000
#> disp         0.00000000  0.0000000  0.000000  0.000000  0.0000000  0.0000000
#> drat         0.00000000  0.0000000  0.000000  0.000000  0.0000000  0.0000000
#> gear         0.00000000  0.0000000  0.000000  0.000000  0.0000000  0.0000000
#> hp           0.00000000  0.0000000  0.000000  0.000000  0.0000000  0.0000000
#> qsec         0.05141792  0.1580864  0.257566  0.350341  0.4368633  0.5175543
#> vs           0.00000000  0.0000000  0.000000  0.000000  0.0000000  0.0000000
#> wt          -5.90000526 -5.8831116 -5.867357 -5.852663 -5.8389604 -5.8261810
#>                 1.0129     0.9447     0.8810     0.8216     0.7662     0.7146
#> (Intercept) 28.0222840 26.7478351 25.5592798 24.4508291 23.4170843 22.4530105
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.5928069  0.6629878  0.7284388  0.7894785  0.8464044  0.8994937
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.8142629 -5.8031480 -5.7927822 -5.7831150 -5.7740994 -5.7656914
#>                 0.6664     0.6215    0.5796    0.5406    0.5041    0.4702
#> (Intercept) 21.5539120 20.7154097 19.933420 19.204133 19.099889 19.099880
#> am           0.0000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.0000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.0000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.0000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.0000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.0000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.0000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.9490049  0.9951792  1.038242  1.078402  1.084142  1.084143
#> vs           0.0000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt          -5.7578501 -5.7505372 -5.743717 -5.737357 -5.736448 -5.736448
#>                0.4385    0.4089    0.3814    0.3557    0.3317    0.3093
#> (Intercept) 19.099880 19.099880 19.099880 19.099880 19.099880 19.099880
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         1.084143  1.084143  1.084143  1.084143  1.084143  1.084143
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -5.736448 -5.736448 -5.736448 -5.736448 -5.736448 -5.736448
#>                0.2885    0.2690    0.2509    0.2340    0.2182    0.2035
#> (Intercept) 19.099880 19.099880 19.099880 19.099880 19.099880 19.099880
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         1.084143  1.084143  1.084143  1.084143  1.084143  1.084143
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -5.736448 -5.736448 -5.736448 -5.736448 -5.736448 -5.736448
#>                0.1898    0.1770    0.1651      0.1540      0.1436      0.1339
#> (Intercept) 19.099880 19.099880 19.099880 12.45775502 12.45564006 12.45359820
#> am           0.000000  0.000000  0.000000 -2.18178870 -2.18207286 -2.18234721
#> carb         0.000000  0.000000  0.000000  1.01449732  1.01467757  1.01485159
#> cyl          0.000000  0.000000  0.000000 -1.45905956 -1.45913997 -1.45921761
#> disp         0.000000  0.000000  0.000000  0.03215433  0.03215858  0.03216268
#> drat         0.000000  0.000000  0.000000  0.00000000  0.00000000  0.00000000
#> gear         0.000000  0.000000  0.000000  0.00000000  0.00000000  0.00000000
#> hp           0.000000  0.000000  0.000000  0.00000000  0.00000000  0.00000000
#> qsec         1.084143  1.084143  1.084143  2.15015675  2.15034621  2.15052912
#> vs           0.000000  0.000000  0.000000 -3.02373358 -3.02401910 -3.02429475
#> wt          -5.736448 -5.736448 -5.736448 -9.19903778 -9.19965434 -9.20024959
#>                  0.1249      0.1165      0.1086      0.1013      0.0945
#> (Intercept) 12.45162691 12.44972377 12.44788641 12.44611256 12.44440002
#> am          -2.18261207 -2.18286778 -2.18311464 -2.18335298 -2.18358307
#> carb         1.01501960  1.01518180  1.01533839  1.01548957  1.01563552
#> cyl         -1.45929256 -1.45936493 -1.45943479 -1.45950223 -1.45956735
#> disp         0.03216663  0.03217045  0.03217414  0.03217771  0.03218114
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         2.15070570  2.15087618  2.15104077  2.15119967  2.15135307
#> vs          -3.02456087 -3.02481779 -3.02506584 -3.02530531 -3.02553650
#> wt          -9.20082426 -9.20137907 -9.20191470 -9.20243182 -9.20293106
#>                  0.0881      0.0822      0.0766      0.0715      0.0666
#> (Intercept) 12.44274668 12.44115048 12.43960946 12.43812171 12.43668538
#> am          -2.18380522 -2.18401968 -2.18422673 -2.18442663 -2.18461961
#> carb         1.01577643  1.01591246  1.01604380  1.01617059  1.01629301
#> cyl         -1.45963021 -1.45969091 -1.45974950 -1.45980607 -1.45986068
#> disp         0.03218446  0.03218767  0.03219076  0.03219375  0.03219663
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         2.15150118  2.15164416  2.15178221  2.15191548  2.15204414
#> vs          -3.02575970 -3.02597518 -3.02618322 -3.02638407 -3.02657797
#> wt          -9.20341305 -9.20387838 -9.20432762 -9.20476133 -9.20518006
#>                  0.0622     0.0580      0.0541      0.0504      0.0470
#> (Intercept) 12.43529869 12.4339599  4.00245223  4.00358419  4.00480734
#> am          -2.18480593 -2.1849858 -2.53772117 -2.53752208 -2.53732189
#> carb         1.01641119  1.0165253  1.02876803  1.02858587  1.02840874
#> cyl         -1.45991341 -1.4599643 -1.70279034 -1.70269188 -1.70259163
#> disp         0.03219942  0.0322021  0.03080472  0.03080049  0.03079644
#> drat         0.00000000  0.0000000  0.38375199  0.38392775  0.38409399
#> gear         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.0000000  0.01607116  0.01607045  0.01606950
#> qsec         2.15216836  2.1522883  2.61087031  2.61070117  2.61052943
#> vs          -3.02676517 -3.0269459 -3.89348542 -3.89330623 -3.89311503
#> wt          -9.20558430 -9.2059746 -9.53308474 -9.53245170 -9.53183249
#>                  0.0438      0.0409      0.0381     0.0356      0.0332
#> (Intercept)  4.00611014  4.00748218  4.00891404  4.0103972  4.01192378
#> am          -2.53712088 -2.53691934 -2.53671756 -2.5365158 -2.53631431
#> carb         1.02823651  1.02806909  1.02790635  1.0277482  1.02759454
#> cyl         -1.70248984 -1.70238677 -1.70228263 -1.7021777 -1.70207209
#> disp         0.03079253  0.03078878  0.03078517  0.0307817  0.03077836
#> drat         0.38425124  0.38440003  0.38454082  0.3846740  0.38480012
#> gear         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> hp           0.01606832  0.01606696  0.01606541  0.0160637  0.01606186
#> qsec         2.61035567  2.61018041  2.61000414  2.6098273  2.60965020
#> vs          -3.89291312 -3.89270171 -3.89248194 -3.8922549 -3.89202146
#> wt          -9.53122706 -9.53063536 -9.53005729 -9.5294928 -9.52894165
#>                  0.0309      0.0288      0.0269      0.0251      0.0234
#> (Intercept)  4.01348688  1.99156192  1.99393973  1.99622231  1.99840741
#> am          -2.53611336 -2.57991484 -2.57967468 -2.57945553 -2.57925634
#> carb         1.02744524  0.96850833  0.96863457  0.96876177  0.96888957
#> cyl         -1.70196612 -1.58367984 -1.58349027 -1.58332238 -1.58317475
#> disp         0.03077515  0.03044379  0.03044852  0.03045312  0.03045758
#> drat         0.38491942  0.39808980  0.39781584  0.39755162  0.39729703
#> gear         0.00000000  0.26583048  0.26584060  0.26583937  0.26582792
#> hp           0.01605988  0.01463708  0.01462787  0.01461934  0.01461144
#> qsec         2.60947329  2.63545405  2.63534394  2.63524543  2.63515804
#> vs          -3.89178266 -3.91630612 -3.91573150 -3.91519927 -3.91470751
#> wt          -9.52840381 -9.44946755 -9.44981763 -9.45017066 -9.45052586
#>                  0.0218      0.0204      0.0190      0.0177      0.0165
#> (Intercept)  2.00049355  2.00247994  2.00436643  2.00615336  2.00784158
#> am          -2.57907611 -2.57891384 -2.57876859 -2.57863943 -2.57852546
#> carb         0.96901764  0.96914569  0.96927347  0.96940071  0.96952721
#> cyl         -1.58304603 -1.58293494 -1.58284025 -1.58276083 -1.58269559
#> disp         0.03046191  0.03046612  0.03047019  0.03047414  0.03047796
#> drat         0.39705193  0.39681616  0.39658955  0.39637191  0.39616303
#> gear         0.26580725  0.26577833  0.26574203  0.26569916  0.26565047
#> hp           0.01460416  0.01459746  0.01459131  0.01458569  0.01458056
#> qsec         2.63508126  2.63501460  2.63495758  2.63490970  2.63487050
#> vs          -3.91425438 -3.91383806 -3.91345677 -3.91310876 -3.91279236
#> wt          -9.45088248 -9.45123983 -9.45159725 -9.45195417 -9.45231001
#>                  0.0154      0.0144      0.0134      0.0125      0.0116
#> (Intercept)  2.00943232  2.01092717  2.01232804  2.01363705  2.01485657
#> am          -2.57842581 -2.57833966 -2.57826619 -2.57820462 -2.57815423
#> carb         0.96965275  0.96977717  0.96990030  0.97002199  0.97014212
#> cyl         -1.58264350 -1.58260360 -1.58257496 -1.58255673 -1.58254810
#> disp         0.03048166  0.03048524  0.03048870  0.03049204  0.03049527
#> drat         0.39596273  0.39577076  0.39558691  0.39541095  0.39524266
#> gear         0.26559665  0.26553832  0.26547608  0.26541046  0.26534195
#> hp           0.01457590  0.01457168  0.01456788  0.01456448  0.01456145
#> qsec         2.63483950  2.63481625  2.63480030  2.63479122  2.63478858
#> vs          -3.91250593 -3.91224787 -3.91201664 -3.91181076 -3.91162880
#> wt          -9.45266427 -9.45301648 -9.45336620 -9.45371302 -9.45405659
#>                  0.0109      0.0101      0.0094      0.0088      0.0082
#> (Intercept)  2.01598913  2.01703741  2.01800420  2.01889236  2.01970484
#> am          -2.57811429 -2.57808412 -2.57806308 -2.57805052 -2.57804586
#> carb         0.97026057  0.97037724  0.97049205  0.97060491  0.97071577
#> cyl         -1.58254829 -1.58255659 -1.58257232 -1.58259484 -1.58262355
#> disp         0.03049839  0.03050140  0.03050431  0.03050711  0.03050981
#> drat         0.39508179  0.39492811  0.39478138  0.39464138  0.39450787
#> gear         0.26527099  0.26519802  0.26512339  0.26504745  0.26497052
#> hp           0.01455877  0.01455643  0.01455439  0.01455264  0.01455116
#> qsec         2.63479198  2.63480102  2.63481531  2.63483450  2.63485821
#> vs          -3.91146936 -3.91133112 -3.91121280 -3.91111317 -3.91103106
#> wt          -9.45439657 -9.45473266 -9.45506459 -9.45539211 -9.45571500
#>                  0.0077      0.0071      0.0067      0.0062      0.0058
#> (Intercept)  2.02044461  2.02111466  2.02171799  2.02225760  2.02273644
#> am          -2.57804853 -2.57805799 -2.57807372 -2.57809524 -2.57812207
#> carb         0.97082457  0.97093126  0.97103580  0.97113817  0.97123834
#> cyl         -1.58265789 -1.58269733 -1.58274139 -1.58278958 -1.58284150
#> disp         0.03051241  0.03051491  0.03051733  0.03051965  0.03052189
#> drat         0.39438062  0.39425941  0.39414400  0.39403419  0.39392976
#> gear         0.26489289  0.26481481  0.26473652  0.26465823  0.26458014
#> hp           0.01454994  0.01454895  0.01454819  0.01454763  0.01454726
#> qsec         2.63488613  2.63491791  2.63495325  2.63499185  2.63503342
#> vs          -3.91096532 -3.91091490 -3.91087875 -3.91085589 -3.91084539
#> wt          -9.45603306 -9.45634612 -9.45665404 -9.45695667 -9.45725391
#>                  0.0054
#> (Intercept)  2.02315746
#> am          -2.57815378
#> carb         0.97133631
#> cyl         -1.58289672
#> disp         0.03052404
#> drat         0.39383048
#> gear         0.26450241
#> hp           0.01454707
#> qsec         2.63507769
#> vs          -3.91084635
#> wt          -9.45754566
#> 
#> $iter
#>   [1]   0   4   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   4   3   3   3   3   3   3   3   3   3   3   3   3   3   3   3   2   1
#>  [37]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1 300   1   1
#>  [55]   1   1   1   1   1   1   1   1   1   1   1   1  89   1   1   1   1   1
#>  [73]   1   1   1  70   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.405683187 5.041358233 4.701587560 4.384716295 4.089201091 3.813602624
#>   [7] 3.556578571 3.316877079 3.093330665 2.884850531 2.690421260 2.509095872
#>  [13] 2.339991208 2.182283632 2.035205019 1.898039012 1.770117536 1.650817538
#>  [19] 1.539557961 1.435796907 1.339029001 1.248782928 1.164619138 1.086127706
#>  [25] 1.012926334 0.944658489 0.880991669 0.821615779 0.766241625 0.714599504
#>  [31] 0.666437889 0.621522206 0.579633689 0.540568319 0.504135823 0.470158756
#>  [37] 0.438471629 0.408920109 0.381360262 0.355657857 0.331687708 0.309333066
#>  [43] 0.288485053 0.269042126 0.250909587 0.233999121 0.218228363 0.203520502
#>  [49] 0.189803901 0.177011754 0.165081754 0.153955796 0.143579691 0.133902900
#>  [55] 0.124878293 0.116461914 0.108612771 0.101292633 0.094465849 0.088099167
#>  [61] 0.082161578 0.076624163 0.071459950 0.066643789 0.062152221 0.057963369
#>  [67] 0.054056832 0.050413582 0.047015876 0.043847163 0.040892011 0.038136026
#>  [73] 0.035565786 0.033168771 0.030933307 0.028848505 0.026904213 0.025090959
#>  [79] 0.023399912 0.021822836 0.020352050 0.018980390 0.017701175 0.016508175
#>  [85] 0.015395580 0.014357969 0.013390290 0.012487829 0.011646191 0.010861277
#>  [91] 0.010129263 0.009446585 0.008809917 0.008216158 0.007662416 0.007145995
#>  [97] 0.006664379 0.006215222 0.005796337 0.005405683
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 51
#> 
#> $loss
#>   [1] 748.42952 630.62735 532.06831 449.98318 381.98112 325.99923 280.25871
#>   [8] 243.22680 213.58381 190.19441 172.08283 158.41123 148.46105 141.61678
#>  [15] 137.35190 135.21667 134.77990 134.77990 129.81858 120.29475 112.34755
#>  [22] 105.74897 100.30224  95.83761  92.20878  89.28977  86.97212  85.16261
#>  [29]  83.78115  82.75902  82.03732  81.56566  81.30094  81.20640  81.20490
#>  [36]  81.20490  81.20490  81.20490  81.20490  81.20490  81.20490  81.20490
#>  [43]  81.20490  81.20490  81.20490  81.20490  81.20490  81.20490  81.20490
#>  [50]  81.20490  81.20490  59.38265  59.38263  59.38262  59.38260  59.38259
#>  [57]  59.38258  59.38257  59.38256  59.38255  59.38254  59.38253  59.38252
#>  [64]  59.38252  59.38251  59.38250  58.05144  58.05143  58.05141  58.05140
#>  [71]  58.05138  58.05137  58.05136  58.05134  58.05133  57.93068  57.93066
#>  [78]  57.93064  57.93062  57.93061  57.93060  57.93059  57.93057  57.93056
#>  [85]  57.93056  57.93055  57.93054  57.93053  57.93053  57.93052  57.93051
#>  [92]  57.93051  57.93050  57.93050  57.93049  57.93049  57.93049  57.93048
#>  [99]  57.93048  57.93048
#> 
#> $linear.predictors
#>      5.4057   5.0414   4.7016   4.3847   4.0892   3.8136   3.5566   3.3169
#> 1  20.70476 20.96794 21.21337 21.44227 21.65574 21.85482 22.04049 22.21364
#> 2  20.70476 20.81563 20.91902 21.01545 21.10538 21.18925 21.26746 21.34040
#> 3  20.70476 21.14712 21.55967 21.94441 22.30322 22.63785 22.94993 23.24097
#> 4  20.70476 20.46622 20.24375 20.03628 19.84279 19.66234 19.49405 19.33710
#> 5  20.70476 20.62748 20.55542 20.48820 20.42552 20.36706 20.31255 20.26170
#> 6  20.70476 20.47816 20.26684 20.06975 19.88595 19.71454 19.55468 19.40559
#> 7  20.70476 20.47816 20.26684 20.06975 19.88595 19.71454 19.55468 19.40559
#> 8  20.70476 20.10187 19.53962 19.01526 18.52624 18.07018 17.64485 17.24819
#> 9  20.70476 20.30495 19.93209 19.58435 19.26005 18.95761 18.67555 18.41250
#> 10 20.70476 20.27509 19.87437 19.50066 19.15214 18.82711 18.52398 18.24128
#> 11 20.70476 19.39708 18.17753 17.04017 15.97947 14.99026 14.06771 13.20735
#> 12 20.70476 21.56821 22.37346 23.12444 23.82481 24.47797 25.08712 25.65520
#> 13 20.70476 21.43680 22.11951 22.75620 23.34999 23.90375 24.42019 24.90183
#> 14 20.70476 20.48115 20.27261 20.07812 19.89674 19.72759 19.56984 19.42271
#> 15 20.70476 20.23925 19.80511 19.40023 19.02264 18.67050 18.34209 18.03581
#> 16 20.70476 20.23626 19.79934 19.39186 19.01185 18.65745 18.32693 18.01869
#> 17 20.70476 21.37708 22.00408 22.58882 23.13416 23.64274 24.11704 24.55938
#> 18 20.70476 21.25463 21.76744 22.24569 22.69171 23.10767 23.49559 23.85737
#> 19 20.70476 21.62913 22.49120 23.29517 24.04495 24.74420 25.39633 26.00450
#> 20 20.70476 20.63943 20.57850 20.52168 20.46869 20.41927 20.37318 20.33019
#> 21 20.70476 20.87834 21.04023 21.19120 21.33200 21.46331 21.58577 21.69997
#>      3.0933   2.8849   2.6904   2.5091    2.3400    2.1823    2.0352    1.8980
#> 1  22.37512 22.52572 22.66617 22.79715 22.919304 23.033226 23.139470 23.238553
#> 2  21.40843 21.47187 21.53104 21.58622 21.637679 21.685671 21.730428 21.772169
#> 3  23.51240 23.76553 24.00161 24.22177 24.427097 24.618584 24.797166 24.963711
#> 4  19.19074 19.05423 18.92693 18.80820 18.697482 18.594222 18.497921 18.408111
#> 5  20.21429 20.17007 20.12883 20.09036 20.054496 20.021045 19.989848 19.960753
#> 6  19.26655 19.13689 19.01596 18.90318 18.798002 18.699913 18.608434 18.523122
#> 7  19.26655 19.13689 19.01596 18.90318 18.798002 18.699913 18.608434 18.523122
#> 8  16.87827 16.53327 16.21153 15.91147 15.631635 15.370660 15.127273 14.900290
#> 9  18.16718 17.93840 17.72503 17.52604 17.340468 17.167399 17.005995 16.855469
#> 10 17.97764 17.73176 17.50246 17.28861 17.089169 16.903173 16.729713 16.567943
#> 11 12.40497 11.65666 10.95879 10.30795  9.700981  9.134917  8.607003  8.114669
#> 12 26.18501 26.67910 27.13990 27.56964 27.970412 28.344177 28.692751 29.017833
#> 13 25.35100 25.76990 26.16057 26.52491 26.864696 27.181580 27.477107 27.752717
#> 14 19.28551 19.15755 19.03821 18.92692 18.823131 18.726335 18.636063 18.551874
#> 15 17.75018 17.48380 17.23537 17.00368 16.787610 16.586101 16.398173 16.222911
#> 16 17.73123 17.46313 17.21311 16.97994 16.762480 16.559679 16.370545 16.194158
#> 17 24.97191 25.35663 25.71543 26.05004 26.362099 26.653127 26.924542 27.177664
#> 18 24.19477 24.50942 24.80287 25.07655 25.331773 25.569799 25.791783 25.998806
#> 19 26.57168 27.10064 27.59395 28.05401 28.483061 28.883198 29.256367 29.604386
#> 20 20.29011 20.25272 20.21786 20.18534 20.155016 20.126735 20.100361 20.075764
#> 21 21.80648 21.90581 21.99844 22.08484 22.165407 22.240546 22.310622 22.375974
#>       1.7701    1.6508    1.5396    1.4358    1.3390   1.2488    1.1646
#> 1  23.308005 23.308005 23.241956 23.104936 22.977150 22.85798 22.746834
#> 2  21.801427 21.801427 21.766249 21.693271 21.625211 21.56174 21.502543
#> 3  25.080449 25.080449 25.122506 25.209755 25.291124 25.36701 25.437778
#> 4  18.345160 18.345160 18.479283 18.757527 19.017018 19.25902 19.484713
#> 5  19.940360 19.940360 20.060973 20.311188 20.544540 20.76217 20.965123
#> 6  18.463323 18.463323 18.498561 18.571663 18.639839 18.70342 18.762715
#> 7  18.463323 18.463323 18.529411 18.666515 18.794378 18.91362 19.024833
#> 8  14.741189 14.741189 14.735281 14.723025 14.711595 14.70093 14.690993
#> 9  16.749960 16.749960 16.751567 16.754900 16.758009 16.76091 16.763612
#> 10 16.454552 16.454552 16.477134 16.523979 16.567668 16.60841 16.646410
#> 11  7.769574  7.769574  7.803097  7.872643  7.937502  7.99799  8.054401
#> 12 29.245694 29.245694 29.277382 29.343121 29.404429 29.46161 29.514928
#> 13 27.945901 27.945901 28.050338 28.266996 28.469052 28.65749 28.833228
#> 14 18.492864 18.492864 18.476643 18.442992 18.411609 18.38234 18.355047
#> 15 16.100063 16.100063 15.989961 15.761549 15.548530 15.34987 15.164596
#> 16 16.070523 16.070523 16.044786 15.991395 15.941602 15.89516 15.851857
#> 17 27.355086 27.355086 27.408919 27.520598 27.624750 27.72188 27.812469
#> 18 26.143916 26.143916 26.086299 25.966770 25.855297 25.75134 25.654382
#> 19 29.848325 29.848325 29.795886 29.687098 29.585642 29.49102 29.402783
#> 20 20.058523 20.058523 19.896174 19.559375 19.245274 18.95234 18.679154
#> 21 22.421782 22.421782 22.307594 22.070706 21.849783 21.64375 21.451601
#>      1.0861    1.0129    0.9447    0.8810    0.8216    0.7662    0.7146
#> 1  22.64318 22.546518 22.456367 22.372292 22.293884 22.220760 22.152565
#> 2  21.44734 21.395852 21.347838 21.303059 21.261298 21.222351 21.186030
#> 3  25.50378 25.565331 25.622735 25.676270 25.726197 25.772760 25.816184
#> 4  19.69520 19.891491 20.074557 20.245285 20.404507 20.552997 20.691480
#> 5  21.15440 21.330924 21.495550 21.649080 21.792262 21.925795 22.050328
#> 6  18.81801 18.869587 18.917683 18.962538 19.004370 19.043383 19.079766
#> 7  19.12855 19.225271 19.315476 19.399602 19.478057 19.551225 19.619462
#> 8  14.68172 14.673075 14.665011 14.657491 14.650477 14.643936 14.637836
#> 9  16.76613 16.768486 16.770679 16.772724 16.774632 16.776411 16.778070
#> 10 16.68185 16.714895 16.745717 16.774461 16.801268 16.826268 16.849583
#> 11  8.10701  8.156073  8.201829  8.244502  8.284299  8.321414  8.356027
#> 12 29.56466 29.611034 29.654286 29.694622 29.732240 29.767323 29.800042
#> 13 28.99712 29.149970 29.292516 29.425456 29.549436 29.665059 29.772891
#> 14 18.32959 18.305851 18.283711 18.263063 18.243807 18.225849 18.209101
#> 15 14.99181 14.830670 14.680389 14.540237 14.409531 14.287634 14.173953
#> 16 15.81147 15.773802 15.738674 15.705913 15.675361 15.646867 15.620294
#> 17 27.89695 27.975737 28.049214 28.117739 28.181645 28.241245 28.296828
#> 18 25.56396 25.479637 25.400995 25.327653 25.259254 25.195465 25.135975
#> 19 29.32049 29.243742 29.172167 29.105415 29.043163 28.985106 28.930962
#> 20 18.42438 18.186771 17.965180 17.758522 17.565793 17.386053 17.218427
#> 21 21.27240 21.105283 20.949427 20.804074 20.668517 20.542097 20.424197
#>       0.6664    0.6215    0.5796    0.5406    0.5041    0.4702    0.4385
#> 1  22.088965 22.029652 21.974337 21.922749 21.915375 21.915375 21.915375
#> 2  21.152156 21.120566 21.091104 21.063628 21.059701 21.059701 21.059701
#> 3  25.856681 25.894449 25.929671 25.962520 25.967215 25.967216 25.967216
#> 4  20.820630 20.941075 21.053403 21.158160 21.173134 21.173135 21.173135
#> 5  22.166468 22.274781 22.375793 22.469998 22.483463 22.483465 22.483465
#> 6  19.113697 19.145342 19.174853 19.202376 19.206310 19.206310 19.206310
#> 7  19.683100 19.742449 19.797798 19.849417 19.856795 19.856796 19.856796
#> 8  14.632147 14.626842 14.621894 14.617279 14.616620 14.616620 14.616620
#> 9  16.779617 16.781060 16.782406 16.783661 16.783841 16.783841 16.783841
#> 10 16.871327 16.891605 16.910517 16.928154 16.930675 16.930675 16.930675
#> 11  8.388307  8.418412  8.446488  8.472671  8.476414  8.476414  8.476414
#> 12 29.830555 29.859011 29.885550 29.910300 29.913838 29.913839 29.913839
#> 13 29.873454 29.967241 30.054706 30.136276 30.147936 30.147937 30.147937
#> 14 18.193482 18.178915 18.165330 18.152661 18.150850 18.150850 18.150850
#> 15 14.067933 13.969059 13.876848 13.790852 13.778560 13.778559 13.778559
#> 16 15.595512 15.572400 15.550846 15.530744 15.527871 15.527871 15.527871
#> 17 28.348665 28.397008 28.442093 28.484139 28.490149 28.490149 28.490150
#> 18 25.080495 25.028753 24.980499 24.935497 24.929065 24.929064 24.929064
#> 19 28.880467 28.833376 28.789458 28.748500 28.742646 28.742645 28.742645
#> 20 17.062098 16.916306 16.780339 16.653536 16.635411 16.635409 16.635409
#> 21 20.314243 20.211700 20.116067 20.026880 20.014132 20.014131 20.014131
#>       0.4089    0.3814    0.3557    0.3317    0.3093    0.2885    0.2690
#> 1  21.915375 21.915375 21.915375 21.915375 21.915375 21.915375 21.915375
#> 2  21.059701 21.059701 21.059701 21.059701 21.059701 21.059701 21.059701
#> 3  25.967216 25.967216 25.967216 25.967216 25.967216 25.967216 25.967216
#> 4  21.173135 21.173135 21.173135 21.173135 21.173135 21.173135 21.173135
#> 5  22.483465 22.483465 22.483465 22.483465 22.483465 22.483465 22.483465
#> 6  19.206310 19.206310 19.206310 19.206310 19.206310 19.206310 19.206310
#> 7  19.856796 19.856796 19.856796 19.856796 19.856796 19.856796 19.856796
#> 8  14.616620 14.616620 14.616620 14.616620 14.616620 14.616620 14.616620
#> 9  16.783841 16.783841 16.783841 16.783841 16.783841 16.783841 16.783841
#> 10 16.930675 16.930675 16.930675 16.930675 16.930675 16.930675 16.930675
#> 11  8.476414  8.476414  8.476414  8.476414  8.476414  8.476414  8.476414
#> 12 29.913839 29.913839 29.913839 29.913839 29.913839 29.913839 29.913839
#> 13 30.147937 30.147937 30.147937 30.147937 30.147937 30.147937 30.147937
#> 14 18.150850 18.150850 18.150850 18.150850 18.150850 18.150850 18.150850
#> 15 13.778559 13.778559 13.778559 13.778559 13.778559 13.778559 13.778559
#> 16 15.527871 15.527871 15.527871 15.527871 15.527871 15.527871 15.527871
#> 17 28.490150 28.490150 28.490150 28.490150 28.490150 28.490150 28.490150
#> 18 24.929064 24.929064 24.929064 24.929064 24.929064 24.929064 24.929064
#> 19 28.742645 28.742645 28.742645 28.742645 28.742645 28.742645 28.742645
#> 20 16.635409 16.635409 16.635409 16.635409 16.635409 16.635409 16.635409
#> 21 20.014131 20.014131 20.014131 20.014131 20.014131 20.014131 20.014131
#>       0.2509    0.2340    0.2182    0.2035    0.1898    0.1770    0.1651
#> 1  21.915375 21.915375 21.915375 21.915375 21.915375 21.915375 21.915375
#> 2  21.059701 21.059701 21.059701 21.059701 21.059701 21.059701 21.059701
#> 3  25.967216 25.967216 25.967216 25.967216 25.967216 25.967216 25.967216
#> 4  21.173135 21.173135 21.173135 21.173135 21.173135 21.173135 21.173135
#> 5  22.483465 22.483465 22.483465 22.483465 22.483465 22.483465 22.483465
#> 6  19.206310 19.206310 19.206310 19.206310 19.206310 19.206310 19.206310
#> 7  19.856796 19.856796 19.856796 19.856796 19.856796 19.856796 19.856796
#> 8  14.616620 14.616620 14.616620 14.616620 14.616620 14.616620 14.616620
#> 9  16.783841 16.783841 16.783841 16.783841 16.783841 16.783841 16.783841
#> 10 16.930675 16.930675 16.930675 16.930675 16.930675 16.930675 16.930675
#> 11  8.476414  8.476414  8.476414  8.476414  8.476414  8.476414  8.476414
#> 12 29.913839 29.913839 29.913839 29.913839 29.913839 29.913839 29.913839
#> 13 30.147937 30.147937 30.147937 30.147937 30.147937 30.147937 30.147937
#> 14 18.150850 18.150850 18.150850 18.150850 18.150850 18.150850 18.150850
#> 15 13.778559 13.778559 13.778559 13.778559 13.778559 13.778559 13.778559
#> 16 15.527871 15.527871 15.527871 15.527871 15.527871 15.527871 15.527871
#> 17 28.490150 28.490150 28.490150 28.490150 28.490150 28.490150 28.490150
#> 18 24.929064 24.929064 24.929064 24.929064 24.929064 24.929064 24.929064
#> 19 28.742645 28.742645 28.742645 28.742645 28.742645 28.742645 28.742645
#> 20 16.635409 16.635409 16.635409 16.635409 16.635409 16.635409 16.635409
#> 21 20.014131 20.014131 20.014131 20.014131 20.014131 20.014131 20.014131
#>      0.1540   0.1436   0.1339   0.1249   0.1165   0.1086   0.1013   0.0945
#> 1  22.01439 22.01441 22.01444 22.01446 22.01448 22.01449 22.01451 22.01453
#> 2  20.87273 20.87270 20.87267 20.87264 20.87261 20.87259 20.87256 20.87254
#> 3  24.57581 24.57554 24.57527 24.57502 24.57478 24.57454 24.57431 24.57409
#> 4  20.57638 20.57633 20.57629 20.57624 20.57620 20.57615 20.57611 20.57607
#> 5  24.00202 24.00211 24.00219 24.00226 24.00234 24.00241 24.00248 24.00255
#> 6  17.82990 17.82979 17.82969 17.82960 17.82950 17.82941 17.82932 17.82924
#> 7  19.11999 19.12000 19.12001 19.12002 19.12003 19.12004 19.12004 19.12005
#> 8  12.66958 12.66932 12.66907 12.66883 12.66859 12.66837 12.66815 12.66794
#> 9  16.22728 16.22727 16.22726 16.22725 16.22724 16.22723 16.22722 16.22721
#> 10 16.62739 16.62743 16.62746 16.62749 16.62752 16.62755 16.62758 16.62760
#> 11 10.38498 10.38512 10.38525 10.38538 10.38550 10.38562 10.38573 10.38584
#> 12 30.84353 30.84372 30.84390 30.84408 30.84425 30.84441 30.84457 30.84472
#> 13 30.62455 30.62466 30.62478 30.62488 30.62499 30.62508 30.62518 30.62527
#> 14 18.18821 18.18826 18.18831 18.18836 18.18841 18.18845 18.18850 18.18854
#> 15 13.90689 13.90689 13.90690 13.90690 13.90690 13.90690 13.90690 13.90690
#> 16 16.96588 16.96604 16.96619 16.96634 16.96649 16.96663 16.96676 16.96689
#> 17 27.80851 27.80841 27.80831 27.80821 27.80812 27.80803 27.80794 27.80786
#> 18 26.55857 26.55856 26.55856 26.55855 26.55855 26.55854 26.55854 26.55853
#> 19 28.92237 28.92240 28.92242 28.92245 28.92247 28.92250 28.92252 28.92254
#> 20 15.96397 15.96393 15.96390 15.96386 15.96383 15.96379 15.96376 15.96373
#> 21 20.11707 20.11711 20.11715 20.11719 20.11723 20.11727 20.11731 20.11735
#>      0.0881   0.0822   0.0766   0.0715   0.0666   0.0622   0.0580   0.0541
#> 1  22.01455 22.01456 22.01458 22.01459 22.01461 22.01462 22.01464 21.55452
#> 2  20.87252 20.87250 20.87247 20.87245 20.87243 20.87241 20.87240 20.58567
#> 3  24.57388 24.57367 24.57347 24.57328 24.57310 24.57292 24.57275 24.55936
#> 4  20.57603 20.57599 20.57596 20.57592 20.57589 20.57586 20.57582 20.40601
#> 5  24.00262 24.00268 24.00274 24.00280 24.00286 24.00291 24.00296 24.09372
#> 6  17.82916 17.82908 17.82900 17.82893 17.82886 17.82879 17.82873 17.63634
#> 7  19.12006 19.12007 19.12007 19.12008 19.12009 19.12009 19.12010 19.20287
#> 8  12.66774 12.66754 12.66735 12.66717 12.66700 12.66683 12.66666 12.66279
#> 9  16.22720 16.22719 16.22718 16.22717 16.22717 16.22716 16.22715 16.42621
#> 10 16.62763 16.62765 16.62768 16.62770 16.62772 16.62775 16.62777 16.99391
#> 11 10.38595 10.38605 10.38615 10.38625 10.38634 10.38643 10.38652 10.34876
#> 12 30.84487 30.84501 30.84515 30.84528 30.84541 30.84553 30.84565 30.83452
#> 13 30.62536 30.62545 30.62553 30.62561 30.62569 30.62577 30.62584 31.10624
#> 14 18.18858 18.18862 18.18866 18.18870 18.18873 18.18877 18.18880 17.84370
#> 15 13.90690 13.90690 13.90690 13.90690 13.90690 13.90690 13.90690 14.27215
#> 16 16.96702 16.96714 16.96725 16.96737 16.96747 16.96758 16.96768 16.61459
#> 17 27.80778 27.80770 27.80763 27.80755 27.80748 27.80742 27.80735 27.74776
#> 18 26.55853 26.55853 26.55852 26.55852 26.55852 26.55851 26.55851 26.78014
#> 19 28.92256 28.92258 28.92260 28.92262 28.92264 28.92266 28.92267 28.71009
#> 20 15.96370 15.96367 15.96364 15.96362 15.96359 15.96356 15.96354 16.26990
#> 21 20.11738 20.11741 20.11745 20.11748 20.11751 20.11754 20.11756 20.15076
#>      0.0504   0.0470   0.0438   0.0409   0.0381   0.0356   0.0332   0.0309
#> 1  21.55452 21.55452 21.55454 21.55455 21.55458 21.55460 21.55463 21.55467
#> 2  20.58574 20.58580 20.58587 20.58594 20.58601 20.58609 20.58616 20.58623
#> 3  24.55956 24.55975 24.55994 24.56012 24.56030 24.56047 24.56064 24.56080
#> 4  20.40596 20.40591 20.40587 20.40583 20.40579 20.40576 20.40573 20.40570
#> 5  24.09368 24.09364 24.09360 24.09355 24.09351 24.09346 24.09342 24.09337
#> 6  17.63649 17.63664 17.63679 17.63694 17.63708 17.63722 17.63737 17.63751
#> 7  19.20291 19.20296 19.20300 19.20304 19.20308 19.20312 19.20316 19.20319
#> 8  12.66304 12.66329 12.66353 12.66375 12.66398 12.66419 12.66439 12.66459
#> 9  16.42622 16.42622 16.42621 16.42621 16.42620 16.42618 16.42617 16.42615
#> 10 16.99388 16.99384 16.99379 16.99375 16.99369 16.99364 16.99358 16.99352
#> 11 10.34861 10.34847 10.34833 10.34819 10.34807 10.34794 10.34782 10.34771
#> 12 30.83446 30.83440 30.83435 30.83430 30.83426 30.83421 30.83417 30.83414
#> 13 31.10615 31.10606 31.10596 31.10586 31.10576 31.10565 31.10554 31.10543
#> 14 17.84367 17.84364 17.84362 17.84361 17.84360 17.84359 17.84359 17.84359
#> 15 14.27217 14.27218 14.27219 14.27219 14.27219 14.27219 14.27219 14.27218
#> 16 16.61442 16.61427 16.61412 16.61399 16.61387 16.61375 16.61365 16.61355
#> 17 27.74785 27.74793 27.74802 27.74810 27.74818 27.74826 27.74834 27.74842
#> 18 26.78024 26.78033 26.78041 26.78048 26.78055 26.78061 26.78066 26.78071
#> 19 28.70991 28.70973 28.70957 28.70942 28.70927 28.70913 28.70900 28.70887
#> 20 16.26991 16.26993 16.26994 16.26995 16.26996 16.26998 16.26999 16.27000
#> 21 20.15062 20.15048 20.15035 20.15021 20.15008 20.14995 20.14982 20.14969
#>      0.0288   0.0269   0.0251   0.0234   0.0218   0.0204   0.0190   0.0177
#> 1  21.50253 21.50277 21.50301 21.50323 21.50343 21.50363 21.50381 21.50398
#> 2  20.56877 20.56886 20.56895 20.56903 20.56910 20.56917 20.56923 20.56928
#> 3  24.49731 24.49717 24.49703 24.49689 24.49676 24.49664 24.49652 24.49640
#> 4  20.41837 20.41852 20.41866 20.41880 20.41892 20.41904 20.41915 20.41926
#> 5  24.14871 24.14851 24.14832 24.14814 24.14797 24.14782 24.14767 24.14754
#> 6  17.69643 17.69643 17.69642 17.69641 17.69639 17.69636 17.69633 17.69629
#> 7  19.27770 19.27763 19.27757 19.27750 19.27743 19.27737 19.27730 19.27723
#> 8  12.67592 12.67568 12.67546 12.67523 12.67502 12.67480 12.67460 12.67439
#> 9  16.41583 16.41569 16.41556 16.41544 16.41533 16.41523 16.41513 16.41504
#> 10 16.99753 16.99734 16.99715 16.99698 16.99682 16.99667 16.99653 16.99641
#> 11 10.30588 10.30604 10.30619 10.30634 10.30648 10.30663 10.30677 10.30691
#> 12 30.73699 30.73715 30.73731 30.73747 30.73762 30.73777 30.73791 30.73804
#> 13 31.09412 31.09398 31.09386 31.09375 31.09364 31.09355 31.09347 31.09340
#> 14 17.89552 17.89579 17.89603 17.89626 17.89647 17.89666 17.89684 17.89700
#> 15 14.04633 14.04609 14.04588 14.04569 14.04552 14.04538 14.04524 14.04513
#> 16 16.62304 16.62343 16.62380 16.62415 16.62449 16.62481 16.62510 16.62539
#> 17 27.71313 27.71313 27.71314 27.71314 27.71314 27.71314 27.71314 27.71313
#> 18 26.89122 26.89083 26.89045 26.89010 26.88976 26.88944 26.88913 26.88884
#> 19 28.71891 28.71915 28.71938 28.71961 28.71982 28.72003 28.72023 28.72042
#> 20 16.43457 16.43462 16.43467 16.43470 16.43473 16.43474 16.43475 16.43476
#> 21 20.14121 20.14118 20.14115 20.14114 20.14113 20.14113 20.14114 20.14115
#>      0.0165   0.0154   0.0144   0.0134   0.0125   0.0116   0.0109   0.0101
#> 1  21.50414 21.50429 21.50443 21.50456 21.50468 21.50479 21.50490 21.50499
#> 2  20.56933 20.56937 20.56941 20.56944 20.56947 20.56949 20.56951 20.56952
#> 3  24.49629 24.49618 24.49608 24.49598 24.49588 24.49578 24.49569 24.49560
#> 4  20.41935 20.41944 20.41953 20.41961 20.41969 20.41976 20.41982 20.41988
#> 5  24.14742 24.14730 24.14720 24.14710 24.14701 24.14693 24.14686 24.14679
#> 6  17.69624 17.69620 17.69614 17.69609 17.69603 17.69597 17.69591 17.69584
#> 7  19.27717 19.27710 19.27703 19.27697 19.27691 19.27684 19.27678 19.27672
#> 8  12.67420 12.67401 12.67382 12.67364 12.67346 12.67329 12.67312 12.67296
#> 9  16.41496 16.41488 16.41481 16.41474 16.41468 16.41462 16.41457 16.41453
#> 10 16.99629 16.99618 16.99608 16.99599 16.99591 16.99584 16.99577 16.99571
#> 11 10.30705 10.30719 10.30732 10.30745 10.30757 10.30770 10.30782 10.30793
#> 12 30.73817 30.73830 30.73842 30.73853 30.73864 30.73875 30.73885 30.73895
#> 13 31.09333 31.09328 31.09323 31.09318 31.09315 31.09312 31.09310 31.09308
#> 14 17.89716 17.89729 17.89742 17.89754 17.89764 17.89774 17.89782 17.89790
#> 15 14.04503 14.04495 14.04488 14.04482 14.04477 14.04474 14.04471 14.04469
#> 16 16.62566 16.62591 16.62615 16.62638 16.62659 16.62679 16.62698 16.62716
#> 17 27.71312 27.71312 27.71311 27.71309 27.71308 27.71307 27.71305 27.71304
#> 18 26.88857 26.88830 26.88806 26.88782 26.88760 26.88739 26.88719 26.88701
#> 19 28.72061 28.72078 28.72095 28.72112 28.72127 28.72142 28.72157 28.72171
#> 20 16.43475 16.43474 16.43473 16.43471 16.43469 16.43466 16.43463 16.43460
#> 21 20.14116 20.14119 20.14121 20.14124 20.14127 20.14131 20.14135 20.14139
#>      0.0094   0.0088   0.0082   0.0077   0.0071   0.0067   0.0062   0.0058
#> 1  21.50508 21.50516 21.50523 21.50530 21.50536 21.50542 21.50547 21.50551
#> 2  20.56953 20.56954 20.56955 20.56955 20.56955 20.56954 20.56954 20.56953
#> 3  24.49552 24.49544 24.49536 24.49528 24.49520 24.49513 24.49506 24.49499
#> 4  20.41994 20.41999 20.42004 20.42008 20.42013 20.42016 20.42020 20.42023
#> 5  24.14673 24.14667 24.14662 24.14658 24.14654 24.14650 24.14647 24.14644
#> 6  17.69577 17.69570 17.69563 17.69556 17.69548 17.69541 17.69534 17.69526
#> 7  19.27666 19.27660 19.27655 19.27649 19.27644 19.27638 19.27633 19.27628
#> 8  12.67280 12.67265 12.67250 12.67236 12.67222 12.67208 12.67195 12.67183
#> 9  16.41449 16.41445 16.41442 16.41439 16.41436 16.41434 16.41432 16.41430
#> 10 16.99566 16.99561 16.99557 16.99554 16.99551 16.99548 16.99547 16.99545
#> 11 10.30805 10.30816 10.30826 10.30837 10.30847 10.30856 10.30866 10.30875
#> 12 30.73904 30.73913 30.73922 30.73930 30.73938 30.73945 30.73952 30.73959
#> 13 31.09307 31.09307 31.09307 31.09307 31.09308 31.09309 31.09311 31.09313
#> 14 17.89797 17.89803 17.89808 17.89813 17.89817 17.89820 17.89823 17.89825
#> 15 14.04469 14.04469 14.04469 14.04470 14.04472 14.04475 14.04477 14.04481
#> 16 16.62732 16.62748 16.62763 16.62776 16.62789 16.62801 16.62813 16.62823
#> 17 27.71302 27.71300 27.71299 27.71297 27.71295 27.71293 27.71291 27.71289
#> 18 26.88683 26.88666 26.88651 26.88636 26.88622 26.88609 26.88597 26.88585
#> 19 28.72184 28.72196 28.72208 28.72220 28.72231 28.72241 28.72251 28.72261
#> 20 16.43456 16.43452 16.43449 16.43444 16.43440 16.43436 16.43431 16.43427
#> 21 20.14144 20.14148 20.14153 20.14158 20.14163 20.14168 20.14174 20.14179
#>      0.0054
#> 1  21.50555
#> 2  20.56952
#> 3  24.49493
#> 4  20.42026
#> 5  24.14642
#> 6  17.69519
#> 7  19.27623
#> 8  12.67170
#> 9  16.41429
#> 10 16.99544
#> 11 10.30884
#> 12 30.73966
#> 13 31.09315
#> 14 17.89827
#> 15 14.04484
#> 16 16.62833
#> 17 27.71287
#> 18 26.88575
#> 19 28.72270
#> 20 16.43422
#> 21 20.14185
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 22.8 18.1 24.4 19.2 17.8 16.4 17.3 15.2 10.4 30.4 33.9 15.2 13.3
#> [16] 19.2 27.3 26.0 30.4 15.8 19.7
#> 
#> $X
#>               am       carb         cyl       disp        drat       gear
#>  [1,]  1.0488088  0.8939982 -0.05643326 -0.4481043  0.39938933  0.2603778
#>  [2,]  1.0488088  0.8939982 -0.05643326 -0.4481043  0.39938933  0.2603778
#>  [3,]  1.0488088 -1.3588772 -1.24153183 -0.9073534  0.30802576  0.2603778
#>  [4,] -0.9534626 -1.3588772 -0.05643326  0.1259570 -1.68370011 -1.1066057
#>  [5,] -0.9534626 -0.6079188 -1.24153183 -0.5655661  0.01566233  0.2603778
#>  [6,] -0.9534626  0.8939982 -0.05643326 -0.3809833  0.43593476  0.2603778
#>  [7,] -0.9534626  0.8939982 -0.05643326 -0.3809833  0.43593476  0.2603778
#>  [8,] -0.9534626  0.1430397  1.12866530  0.5746081 -1.11724597 -1.1066057
#>  [9,] -0.9534626  0.1430397  1.12866530  0.5746081 -1.11724597 -1.1066057
#> [10,] -0.9534626  0.1430397  1.12866530  0.5746081 -1.11724597 -1.1066057
#> [11,] -0.9534626  0.8939982  1.12866530  2.3073902 -1.37306397 -1.1066057
#> [12,]  1.0488088 -0.6079188 -1.24153183 -1.1926177  2.28147891  0.2603778
#> [13,]  1.0488088 -1.3588772 -1.24153183 -1.2332436  0.98411619  0.2603778
#> [14,] -0.9534626 -0.6079188  1.12866530  0.8236624 -0.97106425 -1.1066057
#> [15,] -0.9534626  0.8939982  1.12866530  1.2299212  0.08875318 -1.1066057
#> [16,] -0.9534626 -0.6079188  1.12866530  1.6715068 -1.09897325 -1.1066057
#> [17,]  1.0488088 -1.3588772 -1.24153183 -1.1634730  0.72829819  0.2603778
#> [18,]  1.0488088 -0.6079188 -1.24153183 -0.7987233  1.36784319  1.6273614
#> [19,]  1.0488088 -0.6079188 -1.24153183 -1.0212825  0.16184404  1.6273614
#> [20,]  1.0488088  0.8939982  1.12866530  1.2387529  0.98411619  1.6273614
#> [21,]  1.0488088  2.3959151 -0.05643326 -0.5805800 -0.11224667  1.6273614
#>               hp        qsec         vs         wt
#>  [1,] -0.4542653 -0.82056993 -0.8660254 -0.4815752
#>  [2,] -0.4542653 -0.44229073 -0.8660254 -0.2028726
#>  [3,] -0.7454191  0.63175199  1.1547005 -0.8094606
#>  [4,] -0.5398988  1.71930469  1.1547005  0.4365039
#>  [5,] -1.2763468  1.57069500  1.1547005  0.1414071
#>  [6,] -0.2316182  0.42234744  1.1547005  0.4146449
#>  [7,] -0.2316182  0.82764658  1.1547005  0.4146449
#>  [8,]  0.7446036 -0.18560127 -0.8660254  1.1032042
#>  [9,]  0.7446036 -0.05050156 -0.8660254  0.7316008
#> [10,]  0.7446036  0.21969787 -0.8660254  0.7862484
#> [11,]  1.1727710  0.20618790 -0.8660254  2.3928868
#> [12,] -1.4476138  0.57095712  1.1547005 -1.5799912
#> [13,] -1.2249667  1.50314514  1.1547005 -1.3395420
#> [14,]  0.2308026 -0.25315113 -0.8660254  0.4091802
#> [15,]  1.8578390 -1.52984342 -0.8660254  0.8518254
#> [16,]  0.6589701 -0.42202577 -0.8660254  0.8572902
#> [17,] -1.2078400  0.82764658  1.1547005 -1.2302468
#> [18,] -0.7796725 -0.65845027 -0.8660254 -1.0061918
#> [19,] -0.4028852 -0.52335056  1.1547005 -1.6914723
#> [20,]  2.1832462 -2.14454712 -0.8660254  0.1195481
#> [21,]  0.6589701 -1.46904855 -0.8660254 -0.3176325
#> attr(,"center")
#>  [1]   0.4761905   2.8095238   6.0952381 210.7380952   3.6814286   3.8095238
#>  [7] 136.5238095  17.6747619   0.4285714   3.0606190
#> attr(,"scale")
#>  [1]   0.4994328   1.3316316   1.6876233 113.2283175   0.5472641   0.7315377
#>  [7]  58.3883705   1.4803880   0.4948717   0.9149538
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 42
#> 
#> $lambda.min
#> [1] 0.3093331
#> 
#> $null.dev
#> [1] 35.6395
#> 
#> $Bias
#> [1] 1.704948
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 12.09238