Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 30.26000 29.35694 28.08542 26.94709 26.07199 24.47965 23.16535 22.08909
#>   [9] 21.21535 20.51356 19.95740 19.52425 19.22101 19.02608 18.90701 18.72111
#>  [17] 18.22330 17.63661 17.07802 16.57436 16.20911 16.06267 15.91420 15.74075
#>  [25] 15.60761 15.50832 15.17811 14.97862 14.88685 14.87108 14.78020 14.88667
#>  [33] 14.74926 14.46474 13.90807 13.66071 13.69714 15.04977 15.57505 15.44505
#>  [41] 15.14917 15.20838 17.51517 24.36396 23.98196 24.31201 21.81418 23.08549
#>  [49] 22.97688 23.08320 23.08353 22.49858 22.13253 22.13241 31.27592 35.24914
#>  [57] 35.16477 35.09744 36.22990 36.23238 36.59428 36.59630 36.59827 35.93420
#>  [65] 35.93623 35.80671 35.80851 35.81026 35.91587 34.01062 34.00922 34.00785
#>  [73] 34.00650 34.00517 34.04690 34.00027 33.99892 33.99760 33.99631 33.99505
#>  [81] 33.99382 33.99261 33.99143 33.99028 33.98916 33.98107 33.98039 34.10149
#>  [89] 34.10123 34.17663 34.17637 34.17611 34.17585 34.17558 34.17540 34.27460
#>  [97] 34.27464 34.27465 34.27464 34.27460
#> 
#> $cvse
#>   [1] 12.089127 12.141726 11.847286 11.517263 11.240473 10.441196  9.723583
#>   [8]  9.083315  8.516083  8.018046  7.585612  7.215236  6.900741  6.639222
#>  [15]  6.428526  6.228519  6.015034  5.838620  5.715188  5.602361  5.526764
#>  [22]  5.535668  5.517086  5.464790  5.417103  5.373717  5.142660  4.993005
#>  [29]  4.921915  4.855646  4.813548  4.772135  4.722732  4.505749  4.026849
#>  [36]  3.780061  3.742246  3.801204  3.932552  3.909085  3.831198  3.537207
#>  [43]  3.790440  7.437503  7.457463  7.419247  5.497477  6.393240  6.405480
#>  [50]  6.335101  6.335093  6.259409  6.250810  6.250778  8.849422 10.973613
#>  [57] 10.964394 10.965491 11.669262 11.670501 11.642247 11.643444 11.644607
#>  [64] 11.671269 11.672354 11.686650 11.687685 11.688693 11.666884 10.882062
#>  [71] 10.881748 10.881436 10.881125 10.880816 10.879282 10.875970 10.875672
#>  [78] 10.875374 10.875079 10.874785 10.874494 10.874204 10.873917 10.873632
#>  [85] 10.873349 10.872864 10.872564 10.879506 10.879242 10.872934 10.872680
#>  [92] 10.872427 10.872178 10.871931 10.871680 10.866561 10.866310 10.866063
#>  [99] 10.865820 10.865580
#> 
#> $fold
#>  [1]  6  6  9  7  9  4 10  3  3  5 10  8  1  8  7  4  1  2  1  5  2
#> 
#> $lambda
#>   [1] 4.562358995 4.254871269 3.968107186 3.700670042 3.451257267 3.218654078
#>   [7] 3.001727566 2.799421174 2.610749557 2.434793774 2.270696823 2.117659457
#>  [13] 1.974936297 1.841832200 1.717698875 1.601931719 1.493966883 1.393278515
#>  [19] 1.299376206 1.211802599 1.130131159 1.053964102 0.982930449 0.916684226
#>  [25] 0.854902777 0.797285192 0.743550838 0.693438000 0.646702600 0.603117009
#>  [31] 0.562468941 0.524560417 0.489206801 0.456235900 0.425487127 0.396810719
#>  [37] 0.370067004 0.345125727 0.321865408 0.300172757 0.279942117 0.261074956
#>  [43] 0.243479377 0.227069682 0.211765946 0.197493630 0.184183220 0.171769887
#>  [49] 0.160193172 0.149396688 0.139327852 0.129937621 0.121180260 0.113013116
#>  [55] 0.105396410 0.098293045 0.091668423 0.085490278 0.079728519 0.074355084
#>  [61] 0.069343800 0.064670260 0.060311701 0.056246894 0.052456042 0.048920680
#>  [67] 0.045623590 0.042548713 0.039681072 0.037006700 0.034512573 0.032186541
#>  [73] 0.030017276 0.027994212 0.026107496 0.024347938 0.022706968 0.021176595
#>  [79] 0.019749363 0.018418322 0.017176989 0.016019317 0.014939669 0.013932785
#>  [85] 0.012993762 0.012118026 0.011301312 0.010539641 0.009829304 0.009166842
#>  [91] 0.008549028 0.007972852 0.007435508 0.006934380 0.006467026 0.006031170
#>  [97] 0.005624689 0.005245604 0.004892068 0.004562359
#> 
#> $fit
#> $beta
#>               4.5624     4.2549    3.9681   3.7007    3.4513   3.2187    3.0017
#> (Intercept) 19.86667 21.8899995 23.776967 25.53676 27.177946 28.70852 30.135944
#> am           0.00000  0.0000000  0.000000  0.00000  0.000000  0.00000  0.000000
#> carb         0.00000  0.0000000  0.000000  0.00000  0.000000  0.00000  0.000000
#> cyl          0.00000  0.0000000  0.000000  0.00000  0.000000  0.00000  0.000000
#> disp         0.00000  0.0000000  0.000000  0.00000  0.000000  0.00000  0.000000
#> drat         0.00000  0.0000000  0.000000  0.00000  0.000000  0.00000  0.000000
#> gear         0.00000  0.0000000  0.000000  0.00000  0.000000  0.00000  0.000000
#> hp           0.00000  0.0000000  0.000000  0.00000  0.000000  0.00000  0.000000
#> qsec         0.00000  0.0000000  0.000000  0.00000  0.000000  0.00000  0.000000
#> vs           0.00000  0.0000000  0.000000  0.00000  0.000000  0.00000  0.000000
#> wt           0.00000 -0.6286897 -1.215008 -1.76181 -2.271759 -2.74734 -3.190868
#>                2.7994    2.6107    2.4348    2.2707    2.1177    1.9749
#> (Intercept) 31.467162 32.708661 33.866486 34.946278 35.953295 36.892443
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -3.604504 -3.990262 -4.350022 -4.685534 -4.998435 -5.290247
#>                1.8418    1.7177    1.6019      1.4940     1.3933     1.2994
#> (Intercept) 37.768296 38.585119 39.346891 38.85226263 37.0394821 35.3488045
#> am           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> carb         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> cyl          0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> disp         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> qsec         0.000000  0.000000  0.000000  0.05090614  0.1406236  0.2242979
#> vs           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> wt          -5.562391 -5.816194 -6.052892 -6.18612401 -6.1285323 -6.0748197
#>                 1.2118     1.1301     1.0540     0.9829     0.9167     0.8549
#> (Intercept) 33.7720729 32.3016077 30.9302470 29.6513114 28.4585718 27.3462189
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.3023329  0.3751085  0.4429794  0.5062759  0.5653065  0.6203587
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -6.0247272 -5.9780107 -5.9344428 -5.8938112 -5.8559181 -5.8205788
#>                 0.7973    0.7436     0.6934     0.6467     0.6031     0.5625
#> (Intercept) 26.3088348 25.341367 24.4391031 23.5976489 22.8129059 22.0810519
#> am           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.6717005  0.719582  0.7642365  0.8058814  0.8447196  0.8809402
#> vs           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.7876213 -5.756885 -5.7282202 -5.7014873 -5.6765562 -5.6533053
#>                0.5246    0.4892    0.4562    0.4255    0.3968    0.3701
#> (Intercept) 17.668917 13.279476 10.102371 10.099274 10.097270 10.095973
#> am           1.027162  2.092674  2.863910  2.864565  2.864989  2.865263
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         1.010796  1.136224  1.227008  1.227105  1.227168  1.227208
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -5.135859 -4.605047 -4.220838 -4.220499 -4.220279 -4.220137
#>                0.3451    0.3219    0.3002    0.2799      0.2611     0.2435
#> (Intercept) 10.095134 10.094591 10.094240 10.094013 10.32482896 11.1853558
#> am           2.865441  2.865556  2.865630  2.865678  2.90414623  3.0574924
#> carb         0.000000  0.000000  0.000000  0.000000 -0.02995347 -0.1450356
#> cyl          0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> qsec         1.227234  1.227251  1.227262  1.227270  1.21290700  1.1585478
#> vs           0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> wt          -4.220046 -4.219986 -4.219948 -4.219923 -4.18820860 -4.0634882
#>                 0.2271     0.2118    0.1975    0.1842    0.1718    0.1602
#> (Intercept)  8.4868193  8.4861258  5.153513  5.151215  5.149229  5.147514
#> am           2.5578203  2.5576155  0.000000  0.000000  0.000000  0.000000
#> carb        -0.7466107 -0.7467893 -2.233223 -2.233480 -2.233701 -2.233893
#> cyl          0.0000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.0000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         2.8309671  2.8317984  3.564238  3.564172  3.564114  3.564065
#> gear         0.0000000  0.0000000  4.033521  4.034175  4.034741  4.035230
#> hp           0.0000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.6391733  0.6390209  0.000000  0.000000  0.000000  0.000000
#> vs           0.0000000  0.0000000 -1.632859 -1.633203 -1.633501 -1.633758
#> wt          -2.8119664 -2.8116218 -1.690153 -1.689829 -1.689549 -1.689308
#>                0.1494    0.1393     0.1299      0.1212      0.1130      0.1054
#> (Intercept)  5.146032  5.144752  1.2449879 -13.3559994 -13.3578557 -13.3595101
#> am           0.000000  0.000000  0.0000000   0.0000000   0.0000000   0.0000000
#> carb        -2.234058 -2.234201 -1.9795014  -2.2444264  -2.2447020  -2.2449491
#> cyl          0.000000  0.000000  0.0000000   1.0259376   1.0260570   1.0261609
#> disp         0.000000  0.000000  0.0000000   0.0000000   0.0000000   0.0000000
#> drat         3.564022  3.563985  3.2558603   4.0831051   4.0831417   4.0831781
#> gear         4.035652  4.036017  3.8590237   4.6818871   4.6824920   4.6830325
#> hp           0.000000  0.000000  0.0000000   0.0000000   0.0000000   0.0000000
#> qsec         0.000000  0.000000  0.3655818   0.6029787   0.6029104   0.6028486
#> vs          -1.633980 -1.634172 -2.3441844  -1.6136528  -1.6136092  -1.6135747
#> wt          -1.689099 -1.688918 -2.1217020  -2.5830951  -2.5828465  -2.5826193
#>                  0.0983      0.0917        0.0855       0.0797       0.0744
#> (Intercept) -15.4248147 -15.4242749 -1.698140e+01 -15.57140194 -15.56959732
#> am            1.0666363   1.0670932  9.485283e-01   1.44594905   1.44614413
#> carb         -2.0842669  -2.0839883 -1.977164e+00  -1.43984927  -1.43957148
#> cyl           1.1563332   1.1563599  1.011351e+00   1.18285496   1.18273810
#> disp          0.0000000   0.0000000  9.702371e-03   0.01741292   0.01741464
#> drat          3.8685393   3.8683765  4.022248e+00   3.55996561   3.55972927
#> gear          4.1367496   4.1360711  4.255553e+00   3.83104372   3.83055746
#> hp            0.0000000   0.0000000 -8.313054e-06  -0.01442380  -0.01442574
#> qsec          0.7602029   0.7603143  8.647550e-01   0.94282702   0.94291288
#> vs           -1.1650293  -1.1647075 -9.673498e-01   0.00000000   0.00000000
#> wt           -2.5533062  -2.5535532 -3.449107e+00  -4.23211190  -4.23242506
#>                   0.0693      0.0647       0.0603       0.0562       0.0525
#> (Intercept) -15.56792254 -15.5663682 -15.56492552 -15.56358625 -15.56234271
#> am            1.44632879   1.4465035   1.44666877   1.44682504   1.44697276
#> carb         -1.43930979  -1.4390633  -1.43883110  -1.43861241  -1.43840644
#> cyl           1.18262937   1.1825282   1.18243398   1.18234626   1.18226455
#> disp          0.01741627   0.0174178   0.01741926   0.01742063   0.01742194
#> drat          3.55950759   3.5592996   3.55910437   3.55892106   3.55874886
#> gear          3.83010008   3.8296700   3.82926559   3.82888547   3.82852822
#> hp           -0.01442757  -0.0144293  -0.01443093  -0.01443247  -0.01443393
#> qsec          0.94299453   0.9430722   0.94314598   0.94321613   0.94328279
#> vs            0.00000000   0.0000000   0.00000000   0.00000000   0.00000000
#> wt           -4.23272126  -4.2330015  -4.23326666  -4.23351766  -4.23375530
#>                   0.0489       0.0456       0.0425       0.0397       0.0370
#> (Intercept) -15.56118771 -15.56011459 -15.55911713 -15.55818956 -15.55732654
#> am            1.44711236   1.44724425   1.44736881   1.44748644   1.44759748
#> carb         -1.43821245  -1.43802975  -1.43785769  -1.43769563  -1.43754300
#> cyl           1.18218839   1.18211739   1.18205115   1.18198932   1.18193158
#> disp          0.01742317   0.01742434   0.01742545   0.01742651   0.01742751
#> drat          3.55858705   3.55843492   3.55829184   3.55815722   3.55803049
#> gear          3.82819249   3.82787704   3.82758065   3.82730220   3.82704061
#> hp           -0.01443530  -0.01443660  -0.01443783  -0.01443898  -0.01444008
#> qsec          0.94334611   0.94340624   0.94346334   0.94351754   0.94356897
#> vs            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> wt           -4.23398032  -4.23419343  -4.23439530  -4.23458655  -4.23476774
#>                   0.0345       0.0322       0.0300       0.0280       0.0261
#> (Intercept) -15.55652312 -15.55577470 -15.55507705 -15.55442627 -15.55381874
#> am            1.44770228   1.44780118   1.44789448   1.44798249   1.44806550
#> carb         -1.43739924  -1.43726383  -1.43713627  -1.43701611  -1.43690289
#> cyl           1.18187762   1.18182716   1.18177994   1.18173573   1.18169429
#> disp          0.01742846   0.01742936   0.01743021   0.01743103   0.01743180
#> drat          3.55791115   3.55779873   3.55769276   3.55759285   3.55749862
#> gear          3.82679486   3.82656399   3.82634709   3.82614330   3.82595183
#> hp           -0.01444111  -0.01444208  -0.01444300  -0.01444387  -0.01444469
#> qsec          0.94361777   0.94366404   0.94370792   0.94374952   0.94378893
#> vs            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> wt           -4.23493944  -4.23510215  -4.23525634  -4.23540248  -4.23554099
#>                   0.0243       0.0227       0.0212       0.0197       0.0184
#> (Intercept) -15.55325115 -15.55272043 -15.55222377 -15.55175859 -15.55132250
#> am            1.44814377   1.44821757   1.44828715   1.44835273   1.44841454
#> carb         -1.43679622  -1.43669571  -1.43660099  -1.43651171  -1.43642756
#> cyl           1.18165544   1.18161897   1.18158472   1.18155253   1.18152225
#> disp          0.01743254   0.01743324   0.01743390   0.01743453   0.01743513
#> drat          3.55740969   3.55732575   3.55724648   3.55717160   3.55710085
#> gear          3.82577190   3.82560281   3.82544388   3.82529448   3.82515402
#> hp           -0.01444546  -0.01444619  -0.01444688  -0.01444753  -0.01444815
#> qsec          0.94382628   0.94386165   0.94389515   0.94392686   0.94395687
#> vs            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> wt           -4.23567227  -4.23579669  -4.23591462  -4.23602639  -4.23613231
#>                   0.0172        0.0160       0.0149       0.0139       0.0130
#> (Intercept) -15.55091334 -1.554930e+01 -15.60244400 -15.60430141 -15.60608690
#> am            1.44847280  1.448528e+00   1.35101111   1.35076477   1.35052747
#> carb         -1.43634823 -1.436273e+00  -1.48444957  -1.48473887  -1.48501722
#> cyl           1.18149375  1.181467e+00   1.13554457   1.13561549   1.13568364
#> disp          0.01743570  1.743625e-02   0.01680698   0.01680380   0.01680077
#> drat          3.55703397  3.556971e+00   3.60280835   3.60312317   3.60342437
#> gear          3.82502194  3.824898e+00   3.89086926   3.89129443   3.89170541
#> hp           -0.01444873 -1.444928e-02  -0.01349233  -0.01348935  -0.01348649
#> qsec          0.94398526  9.440121e-01   0.94371037   0.94363653   0.94356557
#> vs            0.00000000 -7.760299e-04  -0.25068966  -0.25096530  -0.25122910
#> wt           -4.23623270 -4.236596e+00  -4.16477109  -4.16432969  -4.16390789
#>                   0.0121       0.0113       0.0105       0.0098       0.0092
#> (Intercept) -15.60780154 -15.60944658 -15.61102346 -15.61253376 -15.61397920
#> am            1.35029890   1.35007874   1.34986670   1.34966250   1.34946586
#> carb         -1.48528496  -1.48554240  -1.48578988  -1.48602771  -1.48625621
#> cyl           1.13574900   1.13581159   1.13587143   1.13592857   1.13598305
#> disp          0.01679789   0.01679514   0.01679252   0.01679002   0.01678764
#> drat          3.60371247   3.60398798   3.60425140   3.60450322   3.60474392
#> gear          3.89210242   3.89248570   3.89285550   3.89321211   3.89355579
#> hp           -0.01348375  -0.01348112  -0.01347860  -0.01347618  -0.01347387
#> qsec          0.94349737   0.94343182   0.94336880   0.94330823   0.94324999
#> vs           -0.25148167  -0.25172359  -0.25195539  -0.25217759  -0.25239066
#> wt           -4.16350479  -4.16311952  -4.16275123  -4.16239914  -4.16206248
#>                   0.0085       0.0080       0.0074       0.0069       0.0065
#> (Intercept) -15.61536158 -15.61668279 -15.61794479 -15.61914956 -15.62029911
#> am            1.34927651   1.34909420   1.34891869   1.34874973   1.34858710
#> carb         -1.48647570  -1.48668649  -1.48688887  -1.48708315  -1.48726961
#> cyl           1.13603494   1.13608429   1.13613118   1.13617569   1.13621790
#> disp          0.01678536   0.01678320   0.01678113   0.01677916   0.01677727
#> drat          3.60497397   3.60519380   3.60540388   3.60560460   3.60579640
#> gear          3.89388686   3.89420563   3.89451240   3.89480751   3.89509128
#> hp           -0.01347164  -0.01346951  -0.01346747  -0.01346551  -0.01346362
#> qsec          0.94319399   0.94314015   0.94308838   0.94303859   0.94299072
#> vs           -0.25259504  -0.25279115  -0.25297938  -0.25316011  -0.25333367
#> wt           -4.16174053  -4.16143260  -4.16113803  -4.16085619  -4.16058649
#>                   0.0060       0.0056       0.0052       0.0049       0.0046
#> (Intercept) -15.62139549 -15.62244072 -15.62343682 -15.62438579 -15.62528961
#> am            1.34843058   1.34827994   1.34813499   1.34799552   1.34786134
#> carb         -1.48744855  -1.48762023  -1.48778493  -1.48794291  -1.48809444
#> cyl           1.13625789   1.13629575   1.13633157   1.13636543   1.13639742
#> disp          0.01677547   0.01677375   0.01677211   0.01677053   0.01676903
#> drat          3.60597966   3.60615477   3.60632209   3.60648197   3.60663476
#> gear          3.89536404   3.89562613   3.89587787   3.89611960   3.89635165
#> hp           -0.01346182  -0.01346009  -0.01345843  -0.01345683  -0.01345530
#> qsec          0.94294468   0.94290040   0.94285782   0.94281687   0.94277749
#> vs           -0.25350040  -0.25366060  -0.25381456  -0.25396256  -0.25410484
#> wt           -4.16032836  -4.16008125  -4.15984467  -4.15961811  -4.15940111
#> 
#> $iter
#>   [1]   0   3   3   3   2   2   2   2   2   2   2   2   2   2   2   2   6   4
#>  [19]   4   4   4   4   4   4   4   4   4   4   4   4   4  51  49  21   1   1
#>  [37]   1   1   1   1   1  21  30  64   1  96   1   1   1   1   1  52  58   1
#>  [55]   1  72   1  78 118   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1  67   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.562358995 4.254871269 3.968107186 3.700670042 3.451257267 3.218654078
#>   [7] 3.001727566 2.799421174 2.610749557 2.434793774 2.270696823 2.117659457
#>  [13] 1.974936297 1.841832200 1.717698875 1.601931719 1.493966883 1.393278515
#>  [19] 1.299376206 1.211802599 1.130131159 1.053964102 0.982930449 0.916684226
#>  [25] 0.854902777 0.797285192 0.743550838 0.693438000 0.646702600 0.603117009
#>  [31] 0.562468941 0.524560417 0.489206801 0.456235900 0.425487127 0.396810719
#>  [37] 0.370067004 0.345125727 0.321865408 0.300172757 0.279942117 0.261074956
#>  [43] 0.243479377 0.227069682 0.211765946 0.197493630 0.184183220 0.171769887
#>  [49] 0.160193172 0.149396688 0.139327852 0.129937621 0.121180260 0.113013116
#>  [55] 0.105396410 0.098293045 0.091668423 0.085490278 0.079728519 0.074355084
#>  [61] 0.069343800 0.064670260 0.060311701 0.056246894 0.052456042 0.048920680
#>  [67] 0.045623590 0.042548713 0.039681072 0.037006700 0.034512573 0.032186541
#>  [73] 0.030017276 0.027994212 0.026107496 0.024347938 0.022706968 0.021176595
#>  [79] 0.019749363 0.018418322 0.017176989 0.016019317 0.014939669 0.013932785
#>  [85] 0.012993762 0.012118026 0.011301312 0.010539641 0.009829304 0.009166842
#>  [91] 0.008549028 0.007972852 0.007435508 0.006934380 0.006467026 0.006031170
#>  [97] 0.005624689 0.005245604 0.004892068 0.004562359
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 31
#> 
#> $loss
#>   [1] 626.60667 542.69332 472.48733 414.01610 365.57658 325.69932 293.11724
#>   [8] 266.73851 245.62309 228.96226 216.06094 206.32232 199.23457 194.35923
#>  [15] 191.32125 189.80028 183.49064 173.77866 165.70963 159.04444 153.57646
#>  [22] 149.12758 145.54437 142.69479 140.46529 138.75834 137.49021 136.58918
#>  [29] 135.99380 135.65158 135.51772 125.89928 120.36328 119.17430 119.17430
#>  [36] 119.17430 119.17430 119.17430 119.17430 119.17430 119.17430 118.60678
#>  [43] 116.88747 107.27708 107.27707  97.10381  97.10380  97.10380  97.10379
#>  [50]  97.10379  97.10378  95.16732  90.56890  90.56889  90.56888  89.49076
#>  [57]  89.49075  87.54669  86.04104  86.04103  86.04102  86.04101  86.04101
#>  [64]  86.04100  86.04099  86.04099  86.04098  86.04098  86.04097  86.04097
#>  [71]  86.04097  86.04097  86.04096  86.04096  86.04096  86.04096  86.04096
#>  [78]  86.04095  86.04095  86.04095  86.04095  86.04069  85.99735  85.99733
#>  [85]  85.99732  85.99731  85.99730  85.99729  85.99728  85.99727  85.99726
#>  [92]  85.99725  85.99725  85.99724  85.99724  85.99723  85.99723  85.99722
#>  [99]  85.99722  85.99721
#> 
#> $linear.predictors
#>      4.5624   4.2549   3.9681   3.7007   3.4513   3.2187   3.0017   2.7994
#> 1  19.86667 20.24283 20.59365 20.92082 21.22594 21.51049 21.77587 22.02336
#> 2  19.86667 20.08252 20.28382 20.47155 20.64664 20.80992 20.96220 21.10421
#> 3  19.86667 20.43144 20.95815 21.44936 21.90746 22.33469 22.73313 23.10471
#> 4  19.86667 19.71473 19.57304 19.44090 19.31766 19.20273 19.09554 18.99558
#> 5  19.86667 19.64558 19.43939 19.24710 19.06776 18.90052 18.74454 18.59908
#> 6  19.86667 19.88448 19.90109 19.91658 19.93103 19.94451 19.95707 19.96879
#> 7  19.86667 19.90963 19.94969 19.98706 20.02190 20.05440 20.08471 20.11297
#> 8  19.86667 19.72731 19.59734 19.47613 19.36309 19.25767 19.15936 19.06767
#> 9  19.86667 19.72731 19.59734 19.47613 19.36309 19.25767 19.15936 19.06767
#> 10 19.86667 19.33123 18.83189 18.36619 17.93188 17.52685 17.14911 16.79683
#> 11 19.86667 19.51355 19.18424 18.87712 18.59070 18.32358 18.07446 17.84214
#> 12 19.86667 18.58938 17.39818 16.28726 15.25121 14.28499 13.38389 12.54352
#> 13 19.86667 20.50688 21.10395 21.66078 22.18008 22.66437 23.11603 23.53725
#> 14 19.86667 20.73635 21.54743 22.30384 23.00927 23.66715 24.28070 24.85290
#> 15 19.86667 20.34028 20.78197 21.19390 21.57806 21.93633 22.27045 22.58206
#> 16 19.86667 19.67701 19.50014 19.33519 19.18135 19.03789 18.90409 18.77931
#> 17 19.86667 19.73045 19.60341 19.48494 19.37445 19.27141 19.17531 19.08569
#> 18 19.86667 19.47269 19.10526 18.76260 18.44303 18.14500 17.86706 17.60784
#> 19 19.86667 20.14853 20.41140 20.65654 20.88517 21.09839 21.29724 21.48269
#> 20 19.86667 19.64558 19.43939 19.24710 19.06776 18.90052 18.74454 18.59908
#> 21 19.86667 20.14224 20.39925 20.63893 20.86245 21.07092 21.26533 21.44664
#>      2.6107   2.4348   2.2707    2.1177    1.9749    1.8418    1.7177    1.6019
#> 1  22.25417 22.46943 22.67018 22.857397 23.031997 23.194831 23.346690 23.488314
#> 2  21.23666 21.36017 21.47537 21.582796 21.682985 21.776421 21.863560 21.944826
#> 3  23.45125 23.77444 24.07584 24.356927 24.619071 24.863548 25.091548 25.304181
#> 4  18.90235 18.81541 18.73433 18.658712 18.588190 18.522422 18.461086 18.403884
#> 5  18.46342 18.33691 18.21892 18.108884 18.006263 17.910559 17.821305 17.738066
#> 6  19.97972 19.98992 19.99942 20.008289 20.016557 20.024268 20.031459 20.038165
#> 7  20.13933 20.16392 20.18684 20.208226 20.228167 20.246763 20.264107 20.280281
#> 8  18.98216 18.90241 18.82804 18.758680 18.693995 18.633670 18.577410 18.524942
#> 9  18.98216 18.90241 18.82804 18.758680 18.693995 18.633670 18.577410 18.524942
#> 10 16.46829 16.16190 15.87615 15.609667 15.361140 15.129363 14.913208 14.711620
#> 11 17.62547 17.42340 17.23496 17.059213 16.895312 16.742457 16.599904 16.466959
#> 12 11.75978 11.02887 10.34722  9.711514  9.118649  8.565742  8.050098  7.569208
#> 13 23.93008 24.29644 24.63810 24.956739 25.253901 25.531035 25.789491 26.030528
#> 14 25.38653 25.88420 26.34832 26.781168 27.184841 27.561308 27.912402 28.239834
#> 15 22.87266 23.14368 23.39644 23.632154 23.851986 24.057001 24.248200 24.426512
#> 16 18.66294 18.55441 18.45320 18.358806 18.270776 18.188679 18.112115 18.040711
#> 17 19.00211 18.92416 18.85147 18.783672 18.720447 18.661482 18.606491 18.555207
#> 18 17.36610 17.14065 16.93040 16.734314 16.551446 16.380901 16.221852 16.073521
#> 19 21.65563 21.81693 21.96735 22.107632 22.238461 22.360472 22.474260 22.580380
#> 20 18.46342 18.33691 18.21892 18.108884 18.006263 17.910559 17.821305 17.738066
#> 21 21.61573 21.77343 21.92049 22.057647 22.185558 22.304848 22.416099 22.519851
#>       1.4940    1.3933    1.2994   1.2118    1.1301    1.0540    0.9829
#> 1  23.482533 23.297391 23.124720 22.96369 22.813506 22.673447 22.542828
#> 2  21.933579 21.813365 21.701248 21.59669 21.499174 21.408233 21.323421
#> 3  25.447818 25.438292 25.429407 25.42112 25.413393 25.406186 25.399465
#> 4  18.477596 18.678169 18.865232 19.03969 19.202385 19.354118 19.495624
#> 5  17.574153 17.388099 17.214577 17.05275 16.901829 16.761079 16.629816
#> 6  20.136650 20.301935 20.456087 20.59985 20.733924 20.858962 20.975572
#> 7  20.531723 20.954885 21.349544 21.71760 22.060859 22.380980 22.679525
#> 8  18.503578 18.530742 18.556076 18.57970 18.601737 18.622286 18.641450
#> 9  18.534122 18.615116 18.690655 18.76110 18.826802 18.888074 18.945216
#> 10 14.560505 14.543206 14.527072 14.51203 14.497993 14.484906 14.472701
#> 11 16.385024 16.404854 16.423348 16.44060 16.456681 16.471682 16.485672
#> 12  7.290404  7.393099  7.488877  7.57820  7.661503  7.739191  7.811644
#> 13 26.233932 26.294652 26.351281 26.40409 26.453347 26.499281 26.542119
#> 14 28.513757 28.592034 28.665038 28.73312 28.796618 28.855834 28.911059
#> 15 24.622099 24.746527 24.862575 24.97080 25.071733 25.165863 25.253648
#> 16 17.935893 17.839368 17.749345 17.66539 17.587091 17.514070 17.445971
#> 17 18.483603 18.420761 18.362152 18.30749 18.256518 18.208979 18.164644
#> 18 15.934565 15.872907 15.815402 15.76177 15.711757 15.665113 15.621612
#> 19 22.505744 22.243113 21.998171 21.76974 21.556700 21.358021 21.172731
#> 20 17.511030 17.213726 16.936447 16.67786 16.436694 16.211785 16.002034
#> 21 22.601692 22.617760 22.632746 22.64672 22.659757 22.671912 22.683249
#>       0.9167    0.8549    0.7973    0.7436    0.6934    0.6467    0.6031
#> 1  22.421012 22.307406 22.201456 22.102648 22.010498 21.924560 21.844413
#> 2  21.244324 21.170559 21.101765 21.037608 20.977775 20.921974 20.869934
#> 3  25.393196 25.387351 25.381899 25.376815 25.372073 25.367651 25.363527
#> 4  19.627593 19.750668 19.865448 19.972493 20.072323 20.165424 20.252251
#> 5  16.507400 16.393234 16.286762 16.187466 16.094863 16.008500 15.927958
#> 6  21.084324 21.185746 21.280332 21.368544 21.450810 21.527532 21.599083
#> 7  22.957949 23.217609 23.459768 23.685607 23.896225 24.092647 24.275832
#> 8  18.659323 18.675991 18.691536 18.706033 18.719553 18.732162 18.743921
#> 9  18.998507 19.048206 19.094556 19.137782 19.178095 19.215691 19.250753
#> 10 14.461319 14.450704 14.440804 14.431572 14.422961 14.414931 14.407443
#> 11 16.498719 16.510887 16.522235 16.532817 16.542687 16.551892 16.560476
#> 12  7.879213  7.942229  8.000997  8.055805  8.106919  8.154588  8.199044
#> 13 26.582070 26.619329 26.654076 26.686481 26.716703 26.744887 26.771172
#> 14 28.962562 29.010594 29.055389 29.097164 29.136125 29.172459 29.206345
#> 15 25.335517 25.411869 25.483074 25.549481 25.611412 25.669169 25.723033
#> 16 17.382461 17.323232 17.267995 17.216480 17.168437 17.123632 17.081847
#> 17 18.123296 18.084735 18.048774 18.015235 17.983958 17.954788 17.927584
#> 18 15.581043 15.543208 15.507924 15.475017 15.444328 15.415708 15.389016
#> 19 20.999930 20.838775 20.688481 20.548316 20.417598 20.295690 20.181999
#> 20 15.806420 15.623989 15.453853 15.295185 15.147209 15.009207 14.880506
#> 21 22.693821 22.703681 22.712876 22.721452 22.729449 22.736908 22.743864
#>       0.5625    0.5246    0.4892   0.4562   0.4255   0.3968   0.3701   0.3451
#> 1  21.769668 21.877828 22.009170 22.10424 22.10428 22.10431 22.10432 22.10434
#> 2  20.821401 21.134230 21.471168 21.71505 21.71523 21.71535 21.71543 21.71548
#> 3  25.359681 25.591797 25.833565 26.00856 26.00871 26.00880 26.00886 26.00890
#> 4  20.333226 20.337138 20.320458 20.30838 20.30841 20.30843 20.30845 20.30846
#> 5  15.852845 15.344908 14.837242 14.46979 14.46944 14.46921 14.46906 14.46897
#> 6  21.665812 21.501445 21.313851 21.17806 21.17798 21.17793 21.17790 21.17788
#> 7  24.446671 24.638187 24.793102 24.90522 24.90541 24.90553 24.90561 24.90566
#> 8  18.754887 18.499127 18.231009 18.03694 18.03678 18.03668 18.03661 18.03657
#> 9  19.283451 19.105604 18.912743 18.77314 18.77304 18.77298 18.77294 18.77291
#> 10 14.400459 14.353820 14.307228 14.27350 14.27347 14.27345 14.27344 14.27343
#> 11 16.568481 16.449696 16.324426 16.23375 16.23368 16.23363 16.23360 16.23358
#> 12  8.240504  8.879768  9.532282 10.00458 10.00500 10.00528 10.00546 10.00557
#> 13 26.795686 27.077384 27.363323 27.57029 27.57048 27.57060 27.57068 27.57073
#> 14 29.237947 29.386615 29.532741 29.63851 29.63861 29.63868 29.63873 29.63876
#> 15 25.773268 25.235050 24.663872 24.25044 24.25012 24.24991 24.24977 24.24968
#> 16 17.042878 16.642820 16.237805 15.94465 15.94438 15.94421 15.94409 15.94402
#> 17 17.902214 17.514010 17.117811 16.83104 16.83078 16.83061 16.83050 16.83043
#> 18 15.364123 15.155609 14.945685 14.79374 14.79360 14.79351 14.79345 14.79341
#> 19 20.075969 20.137086 20.227638 20.29319 20.29319 20.29318 20.29318 20.29318
#> 20 14.760479 15.118683 15.520999 15.81221 15.81239 15.81251 15.81259 15.81264
#> 21 22.750351 23.219194 23.703881 24.05471 24.05501 24.05520 24.05533 24.05541
#>      0.3219   0.3002   0.2799   0.2611   0.2435   0.2271   0.2118    0.1975
#> 1  22.10434 22.10435 22.10435 22.10050 22.08606 22.25241 22.25243 21.827032
#> 2  21.71551 21.71553 21.71554 21.71174 21.69866 21.89329 21.89332 21.396043
#> 3  26.00893 26.00895 26.00896 26.05458 26.23109 26.56851 26.56859 27.222676
#> 4  20.30846 20.30847 20.30847 20.32865 20.40649 18.74836 18.74789 17.377361
#> 5  14.46890 14.46886 14.46884 14.46556 14.44996 14.67357 14.67364 13.728541
#> 6  21.17786 21.17786 21.17785 21.16268 21.10371 21.25316 21.25323 22.948742
#> 7  24.90569 24.90571 24.90573 24.84764 24.62604 23.87036 23.87017 23.836123
#> 8  18.03654 18.03652 18.03651 17.99378 17.82824 18.62147 18.62172 18.879532
#> 9  18.77289 18.77288 18.77287 18.72152 18.52337 19.00498 19.00513 18.879532
#> 10 14.27342 14.27342 14.27342 14.29354 14.37058 14.61497 14.61504 14.617694
#> 11 16.23357 16.23356 16.23356 16.23587 16.24412 15.81394 15.81383 15.107838
#> 12 10.00565 10.00569 10.00572 10.02499 10.10259 10.52462 10.52472  9.891096
#> 13 27.57076 27.57078 27.57080 27.60026 27.71506 28.10675 28.10686 28.245270
#> 14 29.63878 29.63879 29.63880 29.65051 29.69641 29.80430 29.80433 29.361169
#> 15 24.24963 24.24959 24.24957 24.24121 24.20636 24.07315 24.07315 22.409448
#> 16 15.94397 15.94394 15.94392 15.98417 16.13651 15.69180 15.69169 16.675588
#> 17 16.83039 16.83036 16.83034 16.86172 16.98008 17.30974 17.30985 18.209304
#> 18 14.79338 14.79337 14.79336 14.84132 15.02441 15.79887 15.79911 17.266844
#> 19 20.29318 20.29318 20.29318 20.24797 20.07426 18.93112 18.93075 20.142596
#> 20 15.81267 15.81269 15.81270 15.74588 15.49071 14.38659 14.38621 14.038888
#> 21 24.05546 24.05550 24.05552 24.08592 24.20527 25.25805 25.25833 25.138685
#>       0.1842    0.1718    0.1602    0.1494    0.1393   0.1299    0.1212
#> 1  21.826915 21.826815 21.826728 21.826653 21.826588 21.91955 21.630899
#> 2  21.396009 21.395980 21.395955 21.395933 21.395914 21.58324 21.309878
#> 3  27.222891 27.223077 27.223237 27.223376 27.223496 26.77359 26.565828
#> 4  17.377363 17.377365 17.377367 17.377368 17.377369 17.53552 17.511298
#> 5  13.728124 13.727763 13.727452 13.727183 13.726950 13.57170 13.355757
#> 6  22.948993 22.949209 22.949397 22.949558 22.949698 22.93543 22.258952
#> 7  23.836345 23.836538 23.836704 23.836847 23.836972 24.82933 25.050028
#> 8  18.879336 18.879166 18.879019 18.878893 18.878783 18.57336 19.090251
#> 9  18.879336 18.879166 18.879019 18.878893 18.878783 18.79271 19.452038
#> 10 14.617704 14.617714 14.617721 14.617728 14.617734 14.60484 14.677648
#> 11 15.107755 15.107683 15.107621 15.107567 15.107520 15.43949 15.788533
#> 12  9.891242  9.891368  9.891477  9.891571  9.891652  9.87795  9.163263
#> 13 28.245430 28.245569 28.245689 28.245793 28.245882 28.09144 28.333475
#> 14 29.361202 29.361231 29.361255 29.361277 29.361295 29.47888 30.107220
#> 15 22.409065 22.408734 22.408448 22.408202 22.407988 22.63035 21.741096
#> 16 16.675697 16.675791 16.675873 16.675944 16.676005 16.54820 16.757436
#> 17 18.209359 18.209408 18.209450 18.209486 18.209517 18.15554 18.828691
#> 18 17.267037 17.267204 17.267349 17.267474 17.267581 16.96633 17.333060
#> 19 20.142689 20.142769 20.142838 20.142898 20.142949 20.23872 19.714340
#> 20 14.038732 14.038598 14.038481 14.038381 14.038294 13.99286 14.341557
#> 21 25.138775 25.138853 25.138920 25.138978 25.139028 24.66097 24.188755
#>       0.1130    0.1054    0.0983    0.0917    0.0855   0.0797   0.0744   0.0693
#> 1  21.630746 21.630610 21.700333 21.700441 21.584228 22.05057 22.05077 22.05095
#> 2  21.309749 21.309637 21.474953 21.475061 21.188969 21.49936 21.49953 21.49969
#> 3  26.566083 26.566312 26.682439 26.682302 26.714130 26.46285 26.46269 26.46253
#> 4  17.511321 17.511337 17.888169 17.888382 17.743852 18.22878 18.22888 18.22898
#> 5  13.355491 13.355252 13.243354 13.243547 13.754051 13.61322 13.61337 13.61351
#> 6  22.259047 22.259135 21.747875 21.747526 21.721903 21.75707 21.75700 21.75693
#> 7  25.049923 25.049831 24.944360 24.944306 25.235256 24.90062 24.90068 24.90074
#> 8  19.090220 19.090192 18.851100 18.851073 18.585307 18.88512 18.88501 18.88490
#> 9  19.451966 19.451901 19.307222 19.307261 19.104160 19.45082 19.45075 19.45070
#> 10 14.677670 14.677690 14.695289 14.695276 13.976165 13.38081 13.38067 13.38055
#> 11 15.788442 15.788359 15.891870 15.891995 15.495259 15.17382 15.17383 15.17383
#> 12  9.163258  9.163257  9.497443  9.497505  9.770895 10.05134 10.05148 10.05161
#> 13 28.333650 28.333807 28.532374 28.532325 28.512774 28.47957 28.47947 28.47936
#> 14 30.107280 30.107332 30.332814 30.332880 30.632928 30.81019 30.81019 30.81019
#> 15 21.740681 21.740311 21.592827 21.593056 21.734613 21.51117 21.51142 21.51165
#> 16 16.757621 16.757785 16.581720 16.581555 16.554811 16.71257 16.71249 16.71241
#> 17 18.828840 18.828972 18.634368 18.634209 18.652673 18.62232 18.62219 18.62206
#> 18 17.333325 17.333563 17.126664 17.126388 17.672013 17.71330 17.71315 17.71300
#> 19 19.714333 19.714327 19.472567 19.472456 19.265616 19.26645 19.26651 19.26655
#> 20 14.341495 14.341437 14.580389 14.580603 14.986885 14.64204 14.64206 14.64209
#> 21 24.188859 24.188955 24.421869 24.421854 24.313511 23.98798 23.98788 23.98778
#>      0.0647   0.0603   0.0562   0.0525   0.0489   0.0456   0.0425   0.0397
#> 1  22.05113 22.05129 22.05145 22.05160 22.05174 22.05187 22.05199 22.05211
#> 2  21.49983 21.49997 21.50011 21.50023 21.50035 21.50046 21.50056 21.50066
#> 3  26.46238 26.46223 26.46210 26.46197 26.46185 26.46174 26.46163 26.46153
#> 4  18.22908 18.22917 18.22925 18.22933 18.22940 18.22947 18.22954 18.22960
#> 5  13.61364 13.61377 13.61388 13.61399 13.61410 13.61419 13.61428 13.61437
#> 6  21.75686 21.75679 21.75673 21.75667 21.75661 21.75656 21.75651 21.75646
#> 7  24.90079 24.90085 24.90090 24.90094 24.90099 24.90103 24.90107 24.90111
#> 8  18.88480 18.88470 18.88461 18.88453 18.88445 18.88437 18.88430 18.88423
#> 9  19.45064 19.45059 19.45054 19.45050 19.45045 19.45041 19.45038 19.45034
#> 10 13.38042 13.38031 13.38020 13.38010 13.38000 13.37991 13.37982 13.37974
#> 11 15.17384 15.17384 15.17385 15.17386 15.17386 15.17387 15.17387 15.17388
#> 12 10.05174 10.05186 10.05197 10.05207 10.05217 10.05227 10.05235 10.05243
#> 13 28.47927 28.47918 28.47909 28.47901 28.47894 28.47887 28.47880 28.47874
#> 14 30.81019 30.81019 30.81019 30.81020 30.81020 30.81020 30.81021 30.81021
#> 15 21.51186 21.51206 21.51225 21.51243 21.51260 21.51275 21.51290 21.51303
#> 16 16.71234 16.71227 16.71220 16.71214 16.71208 16.71203 16.71198 16.71193
#> 17 18.62194 18.62183 18.62172 18.62163 18.62153 18.62145 18.62137 18.62129
#> 18 17.71287 17.71274 17.71262 17.71251 17.71241 17.71231 17.71222 17.71213
#> 19 19.26660 19.26664 19.26668 19.26671 19.26675 19.26678 19.26681 19.26684
#> 20 14.64211 14.64213 14.64215 14.64217 14.64218 14.64220 14.64222 14.64223
#> 21 23.98768 23.98759 23.98750 23.98742 23.98734 23.98727 23.98720 23.98713
#>      0.0370   0.0345   0.0322   0.0300   0.0280   0.0261   0.0243   0.0227
#> 1  22.05222 22.05232 22.05242 22.05251 22.05259 22.05267 22.05275 22.05282
#> 2  21.50075 21.50084 21.50092 21.50099 21.50107 21.50113 21.50120 21.50126
#> 3  26.46144 26.46135 26.46126 26.46118 26.46110 26.46103 26.46097 26.46090
#> 4  18.22965 18.22971 18.22976 18.22981 18.22986 18.22990 18.22994 18.22998
#> 5  13.61445 13.61452 13.61460 13.61466 13.61473 13.61479 13.61484 13.61490
#> 6  21.75641 21.75636 21.75632 21.75628 21.75624 21.75621 21.75617 21.75614
#> 7  24.90115 24.90119 24.90122 24.90126 24.90129 24.90132 24.90135 24.90138
#> 8  18.88416 18.88410 18.88404 18.88399 18.88394 18.88389 18.88384 18.88380
#> 9  19.45031 19.45027 19.45024 19.45021 19.45019 19.45016 19.45014 19.45011
#> 10 13.37966 13.37959 13.37952 13.37945 13.37939 13.37933 13.37927 13.37922
#> 11 15.17388 15.17389 15.17389 15.17390 15.17390 15.17391 15.17391 15.17391
#> 12 10.05251 10.05258 10.05265 10.05272 10.05278 10.05284 10.05289 10.05294
#> 13 28.47868 28.47862 28.47857 28.47852 28.47848 28.47844 28.47840 28.47836
#> 14 30.81022 30.81023 30.81023 30.81024 30.81024 30.81025 30.81026 30.81026
#> 15 21.51316 21.51328 21.51339 21.51350 21.51360 21.51369 21.51378 21.51386
#> 16 16.71189 16.71185 16.71181 16.71177 16.71174 16.71171 16.71168 16.71165
#> 17 18.62122 18.62116 18.62110 18.62104 18.62099 18.62094 18.62089 18.62084
#> 18 17.71205 17.71198 17.71191 17.71185 17.71178 17.71173 17.71168 17.71163
#> 19 19.26686 19.26689 19.26691 19.26693 19.26695 19.26697 19.26699 19.26700
#> 20 14.64225 14.64227 14.64228 14.64229 14.64231 14.64232 14.64233 14.64235
#> 21 23.98707 23.98701 23.98695 23.98689 23.98684 23.98679 23.98675 23.98670
#>      0.0212   0.0197   0.0184   0.0172   0.0160   0.0149   0.0139   0.0130
#> 1  22.05289 22.05295 22.05301 22.05307 22.05365 22.06520 22.06505 22.06491
#> 2  21.50131 21.50137 21.50142 21.50146 21.50197 21.53166 21.53158 21.53151
#> 3  26.46084 26.46079 26.46073 26.46068 26.46047 26.45044 26.45055 26.45066
#> 4  18.23002 18.23005 18.23008 18.23011 18.22967 18.12863 18.12841 18.12820
#> 5  13.61495 13.61499 13.61504 13.61508 13.61539 13.60677 13.60660 13.60644
#> 6  21.75611 21.75608 21.75605 21.75603 21.75560 21.79563 21.79572 21.79580
#> 7  24.90140 24.90143 24.90145 24.90147 24.90111 24.98322 24.98326 24.98330
#> 8  18.88375 18.88371 18.88367 18.88364 18.88314 18.80920 18.80927 18.80934
#> 9  19.45009 19.45007 19.45005 19.45003 19.44954 19.37543 19.37545 19.37548
#> 10 13.37916 13.37911 13.37907 13.37902 13.37912 13.43848 13.43874 13.43898
#> 11 15.17392 15.17392 15.17392 15.17392 15.17414 15.21249 15.21257 15.21265
#> 12 10.05299 10.05303 10.05308 10.05312 10.05298 10.04278 10.04263 10.04249
#> 13 28.47832 28.47829 28.47826 28.47823 28.47807 28.46230 28.46238 28.46246
#> 14 30.81027 30.81028 30.81028 30.81029 30.81026 30.77839 30.77834 30.77830
#> 15 21.51393 21.51401 21.51407 21.51414 21.51399 21.53484 21.53466 21.53448
#> 16 16.71162 16.71160 16.71158 16.71156 16.71183 16.71054 16.71056 16.71058
#> 17 18.62080 18.62076 18.62073 18.62069 18.62097 18.64014 18.64026 18.64037
#> 18 17.71158 17.71154 17.71150 17.71146 17.71163 17.72062 17.72069 17.72075
#> 19 19.26702 19.26703 19.26705 19.26706 19.26756 19.31860 19.31859 19.31858
#> 20 14.64236 14.64237 14.64238 14.64239 14.64268 14.61452 14.61445 14.61438
#> 21 23.98666 23.98662 23.98658 23.98654 23.98622 23.98011 23.98022 23.98033
#>      0.0121   0.0113   0.0105   0.0098   0.0092   0.0085   0.0080   0.0074
#> 1  22.06477 22.06464 22.06451 22.06439 22.06427 22.06416 22.06405 22.06394
#> 2  21.53144 21.53137 21.53130 21.53123 21.53117 21.53110 21.53104 21.53098
#> 3  26.45077 26.45087 26.45097 26.45106 26.45115 26.45124 26.45132 26.45141
#> 4  18.12800 18.12781 18.12762 18.12744 18.12727 18.12711 18.12695 18.12681
#> 5  13.60628 13.60613 13.60599 13.60585 13.60572 13.60560 13.60548 13.60537
#> 6  21.79588 21.79595 21.79603 21.79610 21.79617 21.79623 21.79630 21.79636
#> 7  24.98334 24.98338 24.98341 24.98345 24.98348 24.98351 24.98354 24.98357
#> 8  18.80940 18.80946 18.80952 18.80957 18.80963 18.80967 18.80972 18.80976
#> 9  19.37550 19.37552 19.37554 19.37556 19.37558 19.37559 19.37560 19.37561
#> 10 13.43921 13.43943 13.43964 13.43985 13.44004 13.44022 13.44040 13.44056
#> 11 15.21273 15.21280 15.21286 15.21293 15.21299 15.21304 15.21310 15.21315
#> 12 10.04235 10.04222 10.04209 10.04197 10.04186 10.04174 10.04164 10.04154
#> 13 28.46253 28.46260 28.46267 28.46274 28.46280 28.46286 28.46292 28.46298
#> 14 30.77826 30.77821 30.77818 30.77814 30.77810 30.77807 30.77804 30.77801
#> 15 21.53432 21.53415 21.53400 21.53385 21.53370 21.53357 21.53343 21.53331
#> 16 16.71060 16.71062 16.71065 16.71067 16.71069 16.71071 16.71073 16.71075
#> 17 18.64048 18.64059 18.64069 18.64079 18.64089 18.64098 18.64106 18.64115
#> 18 17.72081 17.72088 17.72094 17.72100 17.72106 17.72112 17.72118 17.72123
#> 19 19.31857 19.31856 19.31856 19.31855 19.31854 19.31854 19.31853 19.31853
#> 20 14.61432 14.61426 14.61420 14.61415 14.61409 14.61404 14.61400 14.61395
#> 21 23.98043 23.98053 23.98063 23.98072 23.98081 23.98089 23.98097 23.98104
#>      0.0069   0.0065   0.0060   0.0056   0.0052   0.0049   0.0046
#> 1  22.06384 22.06374 22.06365 22.06356 22.06347 22.06339 22.06331
#> 2  21.53092 21.53087 21.53081 21.53076 21.53071 21.53066 21.53062
#> 3  26.45149 26.45156 26.45164 26.45171 26.45177 26.45184 26.45190
#> 4  18.12666 18.12653 18.12639 18.12627 18.12615 18.12603 18.12592
#> 5  13.60526 13.60515 13.60506 13.60496 13.60487 13.60478 13.60470
#> 6  21.79642 21.79648 21.79653 21.79659 21.79664 21.79669 21.79674
#> 7  24.98360 24.98362 24.98365 24.98367 24.98369 24.98372 24.98374
#> 8  18.80980 18.80984 18.80988 18.80991 18.80994 18.80997 18.81000
#> 9  19.37563 19.37563 19.37564 19.37565 19.37566 19.37566 19.37567
#> 10 13.44072 13.44087 13.44102 13.44116 13.44129 13.44142 13.44154
#> 11 15.21319 15.21324 15.21328 15.21332 15.21336 15.21340 15.21343
#> 12 10.04144 10.04135 10.04126 10.04117 10.04109 10.04101 10.04094
#> 13 28.46303 28.46308 28.46313 28.46318 28.46322 28.46327 28.46331
#> 14 30.77798 30.77795 30.77792 30.77790 30.77787 30.77785 30.77782
#> 15 21.53318 21.53307 21.53296 21.53285 21.53275 21.53265 21.53255
#> 16 16.71077 16.71079 16.71081 16.71083 16.71085 16.71087 16.71089
#> 17 18.64123 18.64131 18.64138 18.64146 18.64152 18.64159 18.64165
#> 18 17.72129 17.72134 17.72139 17.72144 17.72149 17.72154 17.72158
#> 19 19.31852 19.31852 19.31852 19.31851 19.31851 19.31851 19.31851
#> 20 14.61391 14.61386 14.61382 14.61379 14.61375 14.61372 14.61368
#> 21 23.98112 23.98119 23.98125 23.98132 23.98138 23.98144 23.98149
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 22.8 18.1 14.3 24.4 22.8 19.2 17.8 16.4 15.2 10.4 32.4 33.9 21.5
#> [16] 15.5 15.2 19.2 19.7 15.0 21.4
#> 
#> $X
#>               am        carb         cyl        disp       drat       gear
#>  [1,]  1.2747549  0.61869102 -0.05643326 -0.50658777  0.7780250  0.5188745
#>  [2,]  1.2747549  0.61869102 -0.05643326 -0.50658777  0.7780250  0.5188745
#>  [3,]  1.2747549 -1.07598439 -1.24153183 -0.98422017  0.6691015  0.5188745
#>  [4,] -0.7844645 -1.07598439 -0.05643326  0.09045273 -1.7054309 -1.0377490
#>  [5,] -0.7844645  0.61869102  1.12866530  1.33045992 -0.7251193 -1.0377490
#>  [6,] -0.7844645 -0.51109258 -1.24153183 -0.62875144  0.3205463  0.5188745
#>  [7,] -0.7844645 -0.51109258 -1.24153183 -0.68294435  0.8215944  0.5188745
#>  [8,] -0.7844645  0.61869102 -0.05643326 -0.43677996  0.8215944  0.5188745
#>  [9,] -0.7844645  0.61869102 -0.05643326 -0.43677996  0.8215944  0.5188745
#> [10,] -0.7844645  0.05379922  1.12866530  0.55706284 -1.0301052 -1.0377490
#> [11,] -0.7844645  0.05379922  1.12866530  0.55706284 -1.0301052 -1.0377490
#> [12,] -0.7844645  0.61869102  1.12866530  2.35920663 -1.3350910 -1.0377490
#> [13,]  1.2747549 -1.07598439 -1.24153183 -1.25334766  1.1701497  0.5188745
#> [14,]  1.2747549 -1.07598439 -1.24153183 -1.32315547  1.4751355  0.5188745
#> [15,] -0.7844645 -1.07598439 -1.24153183 -0.87307879  0.3423310 -1.0377490
#> [16,] -0.7844645 -0.51109258  1.12866530  0.94467991 -1.7054309 -1.0377490
#> [17,] -0.7844645 -0.51109258  1.12866530  0.81608657 -0.8558275 -1.0377490
#> [18,] -0.7844645 -0.51109258  1.12866530  1.69786946 -1.0083205 -1.0377490
#> [19,]  1.2747549  1.74847463 -0.05643326 -0.64436635  0.1680534  2.0754981
#> [20,]  1.2747549  2.87825824  1.12866530  0.78853085 -0.0062242  2.0754981
#> [21,]  1.2747549 -0.51109258 -1.24153183 -0.86481207  1.2355038  0.5188745
#>               hp        qsec         vs          wt
#>  [1,] -0.4758490 -0.90198645 -0.9534626 -0.81556854
#>  [2,] -0.4758490 -0.60123905 -0.9534626 -0.46798640
#>  [3,] -0.7400451  0.25266873  1.0488088 -1.22448870
#>  [4,] -0.5535537  1.11731749  1.0488088  0.32940791
#>  [5,]  1.6221788 -1.23495678 -0.9534626  0.47934530
#>  [6,] -1.2218144  0.99916673  1.0488088 -0.03862024
#>  [7,] -0.7089632  2.55660860  1.0488088 -0.09314293
#>  [8,] -0.2738167  0.08618356  1.0488088  0.30214656
#>  [9,] -0.2738167  0.40841292  1.0488088  0.30214656
#> [10,]  0.6120173 -0.39716046 -0.9534626  1.16087890
#> [11,]  0.6120173 -0.07493111 -0.9534626  0.76558941
#> [12,]  1.0005409 -0.08567209 -0.9534626  2.76929819
#> [13,] -1.1596506  0.71453080  1.0488088 -1.38805676
#> [14,] -1.1751916  0.94546184  1.0488088 -1.88557629
#> [15,] -0.6778813  1.00453722  1.0488088 -1.02684395
#> [16,]  0.1457889 -0.68179639 -0.9534626  0.41119194
#> [17,]  0.1457889 -0.45086536 -0.9534626  0.29533123
#> [18,]  0.5343125 -0.58512759 -0.9534626  0.85418878
#> [19,]  0.5343125 -1.41755341 -0.9534626 -0.61110846
#> [20,]  3.0208640 -1.90089744 -0.9534626  0.47934530
#> [21,] -0.4913899  0.24729824  1.0488088 -0.59747779
#> attr(,"center")
#>  [1]   0.3809524   2.9047619   6.0952381 215.1523810   3.5428571   3.6666667
#>  [7] 140.6190476  18.1395238   0.4761905   3.2183333
#> attr(,"scale")
#>  [1]   0.4856209   1.7702505   1.6876233 108.8703362   0.4590377   0.6424161
#>  [7]  64.3461444   1.8620278   0.4994328   0.7336395
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 36
#> 
#> $lambda.min
#> [1] 0.3968107
#> 
#> $null.dev
#> [1] 29.83841
#> 
#> $Bias
#> [1] 8.439001
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 5.506063