Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 33.595015 31.463713 27.367085 23.578040 20.429361 17.827745 15.692706
#>   [8] 13.954865 12.554458 11.440048 10.567403  9.898520  9.400787  9.046249
#>  [15]  8.810972  8.656497  8.386779  8.207474  8.207474  8.209823  8.182149
#>  [22]  8.082513  8.025740  8.013795  7.841587  7.454066  7.156460  7.162109
#>  [29]  6.956413  6.803226  6.606818  6.450130  6.331763  6.249437  7.351038
#>  [36]  7.334836  7.331882  7.485830  7.805410  7.768447  7.833897  8.329399
#>  [43]  7.894331  8.427263  8.427016  8.760564  8.635125  8.905653  8.906739
#>  [50]  8.906563  8.616484  8.616283  8.998830  9.091825  9.091764  9.302199
#>  [57]  9.072179  9.010955  8.896149  8.895903 10.110401 13.952494 14.043304
#>  [64] 14.043888 14.882506 14.883211 14.809165 14.809718 14.810213 14.810652
#>  [71] 14.837121 14.837499 14.837823 14.838099 14.863497 14.800575 14.757737
#>  [78] 14.823491 14.823610 14.823721 14.809334 14.838239 14.785468 14.844254
#>  [85] 14.854867 14.854905 14.815788 14.815338 14.814906 14.814493 14.814097
#>  [92] 14.813705 14.869408 14.869399 14.883872 14.883943 14.884229 14.889221
#>  [99] 14.889161 14.889098
#> 
#> $cvse
#>   [1] 10.249660 10.282871  9.360884  8.382009  7.511384  6.736115  6.045637
#>   [8]  5.431415  4.886681  4.406201  3.986051  3.623375  3.316091  3.062525
#>  [15]  2.860959  2.704498  2.556469  2.455026  2.455026  2.457972  2.465001
#>  [22]  2.456792  2.437495  2.412601  2.374439  2.302272  2.230736  2.167072
#>  [29]  2.061392  1.996354  1.959933  1.909151  1.866629  1.835669  2.179552
#>  [36]  2.167788  2.159274  2.149034  2.106233  2.109506  2.102005  2.486289
#>  [43]  2.425340  2.469844  2.469843  2.531185  2.537191  2.541225  2.309174
#>  [50]  2.309117  2.199327  2.199319  2.462884  2.462008  2.461916  2.434493
#>  [57]  2.431247  2.431976  2.433948  2.433903  3.433071  5.827018  5.835324
#>  [64]  5.835587  5.877929  5.878041  5.883522  5.883495  5.883409  5.883273
#>  [71]  5.879945  5.879725  5.879475  5.879202  5.875972  5.883133  5.881245
#>  [78]  5.887516  5.887191  5.886863  5.887980  5.884216  5.880043  5.876265
#>  [85]  5.874683  5.874382  5.860210  5.859818  5.859441  5.859078  5.858730
#>  [92]  5.858397  5.851891  5.851576  5.850216  5.849925  5.849679  5.850301
#>  [99]  5.850034  5.849780
#> 
#> $fold
#>  [1] 10  9  3  7 10  2  1  8  4  6  2  5  8  5  7  1  4  9  6  1  3
#> 
#> $lambda
#>   [1] 4.914824175 4.583581475 4.274663424 3.986565416 3.717884250 3.467311295
#>   [7] 3.233626118 3.015690540 2.812443091 2.622893840 2.446119573 2.281259301
#>  [13] 2.127510059 1.984123002 1.850399752 1.725689002 1.609383339 1.500916288
#>  [19] 1.399759554 1.305420445 1.217439476 1.135388130 1.058866770 0.987502693
#>  [25] 0.920948317 0.858879483 0.800993880 0.747009573 0.696663628 0.649710831
#>  [31] 0.605922496 0.565085348 0.527000486 0.491482417 0.458358147 0.427466342
#>  [37] 0.398656542 0.371788425 0.346731129 0.323362612 0.301569054 0.281244309
#>  [43] 0.262289384 0.244611957 0.228125930 0.212751006 0.198412300 0.185039975
#>  [49] 0.172568900 0.160938334 0.150091629 0.139975955 0.130542044 0.121743948
#>  [55] 0.113538813 0.105886677 0.098750269 0.092094832 0.085887948 0.080099388
#>  [61] 0.074700957 0.069666363 0.064971083 0.060592250 0.056508535 0.052700049
#>  [67] 0.049148242 0.045835815 0.042746634 0.039865654 0.037178842 0.034673113
#>  [73] 0.032336261 0.030156905 0.028124431 0.026228938 0.024461196 0.022812593
#>  [79] 0.021275101 0.019841230 0.018503998 0.017256890 0.016093833 0.015009163
#>  [85] 0.013997596 0.013054204 0.012174395 0.011353881 0.010588668 0.009875027
#>  [91] 0.009209483 0.008588795 0.008009939 0.007470096 0.006966636 0.006497108
#>  [97] 0.006059225 0.005650853 0.005270005 0.004914824
#> 
#> $fit
#> $beta
#>               4.9148     4.5836     4.2747     3.9866    3.7179    3.4673
#> (Intercept) 19.60476 21.4657234 23.2012623 24.8198317 26.329315 27.737064
#> am           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> cyl          0.00000 -0.2916432 -0.5636307 -0.8172871 -1.053848 -1.274465
#> disp         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> wt           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#>                3.2336    3.0157    2.8124    2.6229    2.4461    2.2813
#> (Intercept) 29.049936 30.274324 31.416193 32.481103 33.474243 34.400447
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.480214 -1.672096 -1.851045 -2.017934 -2.173575 -2.318727
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                2.1275    1.9841    1.8504   1.7257    1.6094    1.5009
#> (Intercept) 35.264229 36.069795 36.821068 37.52171 38.012812 38.012812
#> am           0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> cyl         -2.454096 -2.580341 -2.698078 -2.80788 -2.884844 -2.884844
#> disp         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#>                1.3998    1.3054    1.2174     1.1354     1.0589    0.9875
#> (Intercept) 38.012812 38.012812 38.012812 37.7326589 37.2327868 36.766239
#> am           0.000000  0.000000  0.000000  0.2258165  0.6287361  1.004794
#> carb         0.000000  0.000000  0.000000  0.0000000  0.0000000  0.000000
#> cyl         -2.884844 -2.884844 -2.884844 -2.8527355 -2.7954453 -2.741975
#> disp         0.000000  0.000000  0.000000  0.0000000  0.0000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.0000000  0.0000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.0000000  0.0000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.0000000  0.0000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.0000000  0.0000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.0000000  0.0000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.0000000  0.0000000  0.000000
#>                0.9209     0.8589     0.8010     0.7470     0.6967    0.6497
#> (Intercept) 36.331135 35.6850014 35.0277162 34.4144259 33.8424681 33.309807
#> am           1.355508  1.9219130  2.5045162  3.0481270  3.5551013  4.027235
#> carb         0.000000 -0.2180713 -0.4706283 -0.7062977 -0.9260843 -1.130728
#> cyl         -2.692107 -2.5260469 -2.3441583 -2.1744371 -2.0161541 -1.868764
#> disp         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> drat         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> gear         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> hp           0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> qsec         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> vs           0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> wt           0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#>                0.6059    0.5651    0.5270    0.4915    0.4584    0.4275
#> (Intercept) 33.186188 33.132227 33.119103 33.118814 33.118672 33.118601
#> am           4.153940  4.220552  4.236774  4.237035  4.237164  4.237227
#> carb        -1.260174 -1.370744 -1.397738 -1.397873 -1.397939 -1.397972
#> cyl         -1.799981 -1.747145 -1.734252 -1.734162 -1.734118 -1.734096
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                0.3987    0.3718    0.3467    0.3234    0.3016    0.2812
#> (Intercept) 33.118567 33.118550 33.118541 27.865245 27.863830 27.862729
#> am           4.237259  4.237274  4.237282  5.246362  5.246797  5.247135
#> carb        -1.397988 -1.397996 -1.398000 -1.390106 -1.390195 -1.390264
#> cyl         -1.734085 -1.734080 -1.734077 -1.103637 -1.103425 -1.103260
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  2.289859  2.290279  2.290607
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                0.2623      0.2446      0.2281      0.2128      0.1984
#> (Intercept) 27.861871 27.17920865 27.17775311 27.17689594 27.17647384
#> am           5.247399  5.09905581  5.09975357  5.10023825  5.10056025
#> carb        -1.390318 -1.16294206 -1.16318512 -1.16337179 -1.16351285
#> cyl         -1.103131 -0.63939889 -0.63929871 -0.63926801 -0.63928486
#> disp         0.000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         0.000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.000000 -0.01901922 -0.01901209 -0.01900608 -0.01900107
#> qsec         0.000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           2.290862  2.11193309  2.11248474  2.11283760  2.11304297
#> wt           0.000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.1850      0.1726      0.1609      0.1501      0.1400
#> (Intercept) 27.17635938 27.17645447 17.80060122 18.48912451 18.48874613
#> am           5.10076010  5.10087007  3.88721488  4.36868509  4.36780885
#> carb        -1.16361749 -1.16369337 -1.49083235 -1.42880003 -1.42876990
#> cyl         -0.63933260 -0.63939882  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         0.00000000  0.00000000  2.17076197  2.76253595  2.76268649
#> gear         0.00000000  0.00000000  0.00000000 -0.90173997 -0.90146668
#> hp          -0.01899695 -0.01899359 -0.02668798 -0.02642599 -0.02643105
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           2.11314136  2.11316433  1.68265801  1.70734497  1.70683505
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.1305      0.1217      0.1135     0.1059      0.0988
#> (Intercept) 18.48855987 18.48850728 18.48854396 13.1489234 13.14589070
#> am           4.36714005  4.36663812  4.36626867  4.8022610  4.80168473
#> carb        -1.42873619 -1.42870177 -1.42866852 -1.4012721 -1.40140161
#> cyl          0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> drat         2.76275859  2.76277638  2.76275814  2.8826761  2.88335618
#> gear        -0.90124214 -0.90106013 -0.90091455 -1.0046901 -1.00465735
#> hp          -0.02643517 -0.02643847 -0.02644108 -0.0228030 -0.02280218
#> qsec         0.00000000  0.00000000  0.00000000  0.2628227  0.26288387
#> vs           1.70645412  1.70617525  1.70597603  1.2699571  1.26938173
#> wt           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#>                  0.0921      0.0859      0.0801     0.0747      0.0697
#> (Intercept) 13.14356023 13.14178844 13.14045815 13.1394741 13.13875938
#> am           4.80114851  4.80066217  4.80023004  4.7998526  4.79952778
#> carb        -1.40150102 -1.40157636 -1.40163263 -1.4016740 -1.40170366
#> cyl          0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> drat         2.88390143  2.88433407  2.88467363  2.8849370  2.88513875
#> gear        -1.00460633 -1.00454520 -1.00447987 -1.0044145 -1.00435182
#> hp          -0.02280204 -0.02280236 -0.02280295 -0.0228037 -0.02280451
#> qsec         0.26292790  0.26295895  0.26298024  0.2629943  0.26300302
#> vs           1.26889141  1.26847811  1.26813335  1.2678487  1.26761591
#> wt           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#>                  0.0650      0.0606      0.0565      0.0527      0.0491
#> (Intercept) 13.13825195 13.13790240 13.13767153 13.13752852 13.13744934
#> am           4.79925188  4.79902035  4.79882820  4.79867044  4.79854226
#> carb        -1.40172446 -1.40173852 -1.40174753 -1.40175285 -1.40175551
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         2.88529101  2.88540403  2.88548625  2.88554458  2.88558464
#> gear        -1.00429361 -1.00424085 -1.00419394 -1.00415292 -1.00411756
#> hp          -0.02280534 -0.02280613 -0.02280686 -0.02280752 -0.02280811
#> qsec         0.26300793  0.26301013  0.26301047  0.26300960  0.26300797
#> vs           1.26742749  1.26727647  1.26715669  1.26706271  1.26698983
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.0458      0.0427      0.0399      0.0372      0.0347
#> (Intercept) 13.13741547 13.13741278 13.13743066 13.13746122 13.13749877
#> am           4.79843918  4.79835717  4.79829262  4.79824242  4.79820386
#> carb        -1.40175631 -1.40175587 -1.40175462 -1.40175291 -1.40175097
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         2.88561090  2.88562693  2.88563552  2.88563883  2.88563853
#> gear        -1.00408747 -1.00406216 -1.00404110 -1.00402376 -1.00400963
#> hp          -0.02280862 -0.02280905 -0.02280942 -0.02280973 -0.02280999
#> qsec         0.26300592  0.26300369  0.26300144  0.26299928  0.26299727
#> vs           1.26693404  1.26689194  1.26686070  1.26683798  1.26682187
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                 0.0323      0.0302     0.0281      0.0262     0.0245
#> (Intercept) 13.1375392 13.13757980 13.1376187 13.13765469 13.1376873
#> am           4.7981747  4.79815289  4.7981370  4.79812564  4.7981178
#> carb        -1.4017490 -1.40174700 -1.4017452 -1.40174348 -1.4017420
#> cyl          0.0000000  0.00000000  0.0000000  0.00000000  0.0000000
#> disp         0.0000000  0.00000000  0.0000000  0.00000000  0.0000000
#> drat         2.8856359  2.88563176  2.8856269  2.88562172  2.8856166
#> gear        -1.0039982 -1.00398912 -1.0039819 -1.00397630 -1.0039720
#> hp          -0.0228102 -0.02281037 -0.0228105 -0.02281061 -0.0228107
#> qsec         0.2629955  0.26299385  0.2629924  0.26299124  0.2629902
#> vs           1.2668108  1.26680356  1.2667991  1.26679677  1.2667959
#> wt           0.0000000  0.00000000  0.0000000  0.00000000  0.0000000
#>                  0.0228        0.0213       0.0198       0.0185       0.0173
#> (Intercept) 13.13771611  1.313907e+01 13.617157164 13.615994734 13.614922765
#> am           4.79811257  4.797925e+00  4.808141509  4.807364017  4.806663746
#> carb        -1.40174066 -1.401687e+00 -1.410216944 -1.410317365 -1.410402816
#> cyl          0.00000000  0.000000e+00  0.000000000  0.000000000  0.000000000
#> disp         0.00000000 -2.243683e-06 -0.001730612 -0.001733983 -0.001736642
#> drat         2.88561171  2.885376e+00  2.777859333  2.778272078  2.778649841
#> gear        -1.00396865 -1.003939e+00 -1.021638668 -1.021561119 -1.021483463
#> hp          -0.02281076 -2.280912e-02 -0.020385691 -0.020382495 -0.020380234
#> qsec         0.26298936  2.629673e-01  0.267083320  0.267112610  0.267136434
#> vs           1.26679595  1.266813e+00  1.226138818  1.225464023  1.224873409
#> wt           0.00000000  0.000000e+00  0.000000000  0.000000000  0.000000000
#>                   0.0161       0.0150       0.0140       0.0131       0.0122
#> (Intercept) 14.262755977 14.264335627 14.265827506 14.267235803 14.268564543
#> am           4.829485062  4.829146818  4.828824962  4.828518851  4.828227860
#> carb        -1.525451165 -1.525649947 -1.525836952 -1.526012785 -1.526178024
#> cyl          0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> disp        -0.004178447 -0.004181983 -0.004185292 -0.004188389 -0.004191283
#> drat         2.865872701  2.866179301  2.866469547  2.866744181  2.867003917
#> gear        -1.033996231 -1.033936769 -1.033879756 -1.033825126 -1.033772810
#> hp          -0.019429058 -0.019430350 -0.019431618 -0.019432860 -0.019434074
#> qsec         0.195835701  0.195668683  0.195510858  0.195361792  0.195221068
#> vs           1.231980161  1.231894241  1.231813228  1.231736880  1.231664964
#> wt           0.338305147  0.338857096  0.339376700  0.339865601  0.340325368
#>                   0.0114       0.0106       0.0099       0.0092       0.0086
#> (Intercept) 14.269817597 14.270998683 14.272111370 14.273159082 14.274145101
#> am           4.827951376  4.827688805  4.827439568  4.827203101  4.826978859
#> carb        -1.526333221 -1.526478907 -1.526615587 -1.526743742 -1.526863834
#> cyl          0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> disp        -0.004193987 -0.004196512 -0.004198867 -0.004201063 -0.004203108
#> drat         2.867249445  2.867481428  2.867700509  2.867907302  2.868102401
#> gear        -1.033722741 -1.033674849 -1.033629065 -1.033585322 -1.033543551
#> hp          -0.019435260 -0.019436416 -0.019437541 -0.019438636 -0.019439700
#> qsec         0.195088284  0.194963055  0.194845010  0.194733793  0.194629064
#> vs           1.231597259  1.231533551  1.231473635  1.231417317  1.231364408
#> wt           0.340757505  0.341163453  0.341544587  0.341902221  0.342237611
#>                   0.0080       0.0075       0.0070       0.0065       0.0061
#> (Intercept) 14.275072572 14.275944507 14.276763787 14.277533168 14.278255285
#> am           4.826766311  4.826564943  4.826374260  4.826193782  4.826023043
#> carb        -1.526976302 -1.527081565 -1.527180023 -1.527272055 -1.527358024
#> cyl          0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> disp        -0.004205012 -0.004206783 -0.004208429 -0.004209957 -0.004211374
#> drat         2.868286375  2.868459771  2.868623115  2.868776910  2.868921639
#> gear        -1.033503686 -1.033465658 -1.033429402 -1.033394855 -1.033361951
#> hp          -0.019440733 -0.019441734 -0.019442703 -0.019443641 -0.019444548
#> qsec         0.194530494  0.194437770  0.194350592  0.194268672  0.194191735
#> vs           1.231314731  1.231268114  1.231224393  1.231183413  1.231145024
#> wt           0.342551958  0.342846403  0.343122039  0.343379905  0.343620991
#>                   0.0057       0.0053      0.0049
#> (Intercept) 14.278932656 14.279567682 14.28016266
#> am           4.825861597  4.825709011  4.82556487
#> carb        -1.527438274 -1.527513132 -1.52758291
#> cyl          0.000000000  0.000000000  0.00000000
#> disp        -0.004212688 -0.004213905 -0.00421503
#> drat         2.869057765  2.869185731  2.86930596
#> gear        -1.033330630 -1.033300829 -1.03327249
#> hp          -0.019445424 -0.019446269 -0.01944708
#> qsec         0.194119520  0.194051773  0.19398826
#> vs           1.231109086  1.231075462  1.23104402
#> wt           0.343846240  0.344056550  0.34425277
#> 
#> $iter
#>   [1]  0  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  8  7  7  7
#>  [26] 23 17 17 17 16 11 11  6  1  1  1  1  1  1 57  1  1  1 54  1  1  1  1  1 53
#>  [51] 49  1  1  1  1 37  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#>  [76]  1  1  1  2 35  1  1 82  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> 
#> $lambda
#>   [1] 4.914824175 4.583581475 4.274663424 3.986565416 3.717884250 3.467311295
#>   [7] 3.233626118 3.015690540 2.812443091 2.622893840 2.446119573 2.281259301
#>  [13] 2.127510059 1.984123002 1.850399752 1.725689002 1.609383339 1.500916288
#>  [19] 1.399759554 1.305420445 1.217439476 1.135388130 1.058866770 0.987502693
#>  [25] 0.920948317 0.858879483 0.800993880 0.747009573 0.696663628 0.649710831
#>  [31] 0.605922496 0.565085348 0.527000486 0.491482417 0.458358147 0.427466342
#>  [37] 0.398656542 0.371788425 0.346731129 0.323362612 0.301569054 0.281244309
#>  [43] 0.262289384 0.244611957 0.228125930 0.212751006 0.198412300 0.185039975
#>  [49] 0.172568900 0.160938334 0.150091629 0.139975955 0.130542044 0.121743948
#>  [55] 0.113538813 0.105886677 0.098750269 0.092094832 0.085887948 0.080099388
#>  [61] 0.074700957 0.069666363 0.064971083 0.060592250 0.056508535 0.052700049
#>  [67] 0.049148242 0.045835815 0.042746634 0.039865654 0.037178842 0.034673113
#>  [73] 0.032336261 0.030156905 0.028124431 0.026228938 0.024461196 0.022812593
#>  [79] 0.021275101 0.019841230 0.018503998 0.017256890 0.016093833 0.015009163
#>  [85] 0.013997596 0.013054204 0.012174395 0.011353881 0.010588668 0.009875027
#>  [91] 0.009209483 0.008588795 0.008009939 0.007470096 0.006966636 0.006497108
#>  [97] 0.006059225 0.005650853 0.005270005 0.004914824
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 25
#> 
#> $loss
#>   [1] 656.98952 559.60990 478.13736 410.28276 354.06975 307.79305 269.98225
#>   [8] 239.37031 214.86632 195.53178 180.56007 169.25862 161.03344 155.37571
#>  [15] 151.85020 150.08514 149.72409 149.72409 149.72409 149.72409 149.72409
#>  [22] 144.63106 136.41371 129.75025 124.41127 108.17105  93.84360  83.83182
#>  [29]  77.41562  73.97578  72.96102  72.61069  72.59643  72.59643  72.59643
#>  [36]  72.59643  72.59643  72.59643  72.59643  64.79812  64.79811  64.79811
#>  [43]  64.79811  58.65089  58.65088  58.65087  58.65087  58.65087  58.65087
#>  [50]  55.16829  53.27645  53.27644  53.27643  53.27643  53.27643  52.20811
#>  [57]  52.20811  52.20810  52.20810  52.20809  52.20809  52.20809  52.20809
#>  [64]  52.20809  52.20809  52.20809  52.20809  52.20809  52.20809  52.20809
#>  [71]  52.20809  52.20809  52.20809  52.20809  52.20809  52.20809  52.20809
#>  [78]  52.20809  52.20785  52.11409  52.11408  52.11407  52.00266  52.00265
#>  [85]  52.00264  52.00262  52.00262  52.00261  52.00260  52.00259  52.00259
#>  [92]  52.00258  52.00258  52.00257  52.00257  52.00257  52.00256  52.00256
#>  [99]  52.00256  52.00256
#> 
#> $linear.predictors
#>      4.9148   4.5836   4.2747   3.9866   3.7179   3.4673   3.2336   3.0157
#> 1  19.60476 19.71586 19.81948 19.91611 20.00623 20.09027 20.16865 20.24175
#> 2  19.60476 19.71586 19.81948 19.91611 20.00623 20.09027 20.16865 20.24175
#> 3  19.60476 19.71586 19.81948 19.91611 20.00623 20.09027 20.16865 20.24175
#> 4  19.60476 19.13258 18.69222 18.28154 17.89853 17.54134 17.20823 16.89756
#> 5  19.60476 20.29915 20.94674 21.55068 22.11392 22.63920 23.12908 23.58594
#> 6  19.60476 19.71586 19.81948 19.91611 20.00623 20.09027 20.16865 20.24175
#> 7  19.60476 19.71586 19.81948 19.91611 20.00623 20.09027 20.16865 20.24175
#> 8  19.60476 19.13258 18.69222 18.28154 17.89853 17.54134 17.20823 16.89756
#> 9  19.60476 19.13258 18.69222 18.28154 17.89853 17.54134 17.20823 16.89756
#> 10 19.60476 19.13258 18.69222 18.28154 17.89853 17.54134 17.20823 16.89756
#> 11 19.60476 19.13258 18.69222 18.28154 17.89853 17.54134 17.20823 16.89756
#> 12 19.60476 20.29915 20.94674 21.55068 22.11392 22.63920 23.12908 23.58594
#> 13 19.60476 20.29915 20.94674 21.55068 22.11392 22.63920 23.12908 23.58594
#> 14 19.60476 20.29915 20.94674 21.55068 22.11392 22.63920 23.12908 23.58594
#> 15 19.60476 19.13258 18.69222 18.28154 17.89853 17.54134 17.20823 16.89756
#> 16 19.60476 19.13258 18.69222 18.28154 17.89853 17.54134 17.20823 16.89756
#> 17 19.60476 19.13258 18.69222 18.28154 17.89853 17.54134 17.20823 16.89756
#> 18 19.60476 19.13258 18.69222 18.28154 17.89853 17.54134 17.20823 16.89756
#> 19 19.60476 20.29915 20.94674 21.55068 22.11392 22.63920 23.12908 23.58594
#> 20 19.60476 20.29915 20.94674 21.55068 22.11392 22.63920 23.12908 23.58594
#> 21 19.60476 19.13258 18.69222 18.28154 17.89853 17.54134 17.20823 16.89756
#>      2.8124   2.6229   2.4461   2.2813   2.1275   1.9841   1.8504   1.7257
#> 1  20.30992 20.37350 20.43279 20.48809 20.53966 20.58775 20.63260 20.67443
#> 2  20.30992 20.37350 20.43279 20.48809 20.53966 20.58775 20.63260 20.67443
#> 3  20.30992 20.37350 20.43279 20.48809 20.53966 20.58775 20.63260 20.67443
#> 4  16.60783 16.33763 16.08564 15.85063 15.63146 15.42707 15.23645 15.05867
#> 5  24.01201 24.40937 24.77994 25.12554 25.44785 25.74843 26.02876 26.29019
#> 6  20.30992 20.37350 20.43279 20.48809 20.53966 20.58775 20.63260 20.67443
#> 7  20.30992 20.37350 20.43279 20.48809 20.53966 20.58775 20.63260 20.67443
#> 8  16.60783 16.33763 16.08564 15.85063 15.63146 15.42707 15.23645 15.05867
#> 9  16.60783 16.33763 16.08564 15.85063 15.63146 15.42707 15.23645 15.05867
#> 10 16.60783 16.33763 16.08564 15.85063 15.63146 15.42707 15.23645 15.05867
#> 11 16.60783 16.33763 16.08564 15.85063 15.63146 15.42707 15.23645 15.05867
#> 12 24.01201 24.40937 24.77994 25.12554 25.44785 25.74843 26.02876 26.29019
#> 13 24.01201 24.40937 24.77994 25.12554 25.44785 25.74843 26.02876 26.29019
#> 14 24.01201 24.40937 24.77994 25.12554 25.44785 25.74843 26.02876 26.29019
#> 15 16.60783 16.33763 16.08564 15.85063 15.63146 15.42707 15.23645 15.05867
#> 16 16.60783 16.33763 16.08564 15.85063 15.63146 15.42707 15.23645 15.05867
#> 17 16.60783 16.33763 16.08564 15.85063 15.63146 15.42707 15.23645 15.05867
#> 18 16.60783 16.33763 16.08564 15.85063 15.63146 15.42707 15.23645 15.05867
#> 19 24.01201 24.40937 24.77994 25.12554 25.44785 25.74843 26.02876 26.29019
#> 20 24.01201 24.40937 24.77994 25.12554 25.44785 25.74843 26.02876 26.29019
#> 21 16.60783 16.33763 16.08564 15.85063 15.63146 15.42707 15.23645 15.05867
#>      1.6094   1.5009   1.3998   1.3054   1.2174   1.1354   1.0589   0.9875
#> 1  20.70375 20.70375 20.70375 20.70375 20.70375 20.84206 21.08885 21.31919
#> 2  20.70375 20.70375 20.70375 20.70375 20.70375 20.84206 21.08885 21.31919
#> 3  20.70375 20.70375 20.70375 20.70375 20.70375 20.61625 20.46011 20.31439
#> 4  14.93406 14.93406 14.93406 14.93406 14.93406 14.91078 14.86922 14.83044
#> 5  26.47344 26.47344 26.47344 26.47344 26.47344 26.32172 26.05101 25.79834
#> 6  20.70375 20.70375 20.70375 20.70375 20.70375 20.61625 20.46011 20.31439
#> 7  20.70375 20.70375 20.70375 20.70375 20.70375 20.61625 20.46011 20.31439
#> 8  14.93406 14.93406 14.93406 14.93406 14.93406 14.91078 14.86922 14.83044
#> 9  14.93406 14.93406 14.93406 14.93406 14.93406 14.91078 14.86922 14.83044
#> 10 14.93406 14.93406 14.93406 14.93406 14.93406 14.91078 14.86922 14.83044
#> 11 14.93406 14.93406 14.93406 14.93406 14.93406 14.91078 14.86922 14.83044
#> 12 26.47344 26.47344 26.47344 26.47344 26.47344 26.54753 26.67974 26.80314
#> 13 26.47344 26.47344 26.47344 26.47344 26.47344 26.54753 26.67974 26.80314
#> 14 26.47344 26.47344 26.47344 26.47344 26.47344 26.32172 26.05101 25.79834
#> 15 14.93406 14.93406 14.93406 14.93406 14.93406 14.91078 14.86922 14.83044
#> 16 14.93406 14.93406 14.93406 14.93406 14.93406 14.91078 14.86922 14.83044
#> 17 14.93406 14.93406 14.93406 14.93406 14.93406 14.91078 14.86922 14.83044
#> 18 14.93406 14.93406 14.93406 14.93406 14.93406 14.91078 14.86922 14.83044
#> 19 26.47344 26.47344 26.47344 26.47344 26.47344 26.54753 26.67974 26.80314
#> 20 26.47344 26.47344 26.47344 26.47344 26.47344 26.54753 26.67974 26.80314
#> 21 14.93406 14.93406 14.93406 14.93406 14.93406 15.13659 15.49796 15.83524
#>      0.9209   0.8589   0.8010   0.7470   0.6967   0.6497   0.6059   0.5651
#> 1  21.53400 21.57835 21.58477 21.59074 21.59631 21.60155 21.49955 21.38693
#> 2  21.53400 21.57835 21.58477 21.59074 21.59631 21.60155 21.49955 21.38693
#> 3  20.17849 20.31065 20.49214 20.66151 20.81946 20.96650 21.12613 21.27861
#> 4  14.79428 14.60434 14.39194 14.19374 14.00890 13.83678 13.74565 13.67209
#> 5  25.56271 25.14467 24.70983 24.30408 23.92568 23.57330 23.46592 23.40216
#> 6  20.17849 19.65643 19.08025 18.54261 18.04121 17.57431 17.34561 17.16638
#> 7  20.17849 19.65643 19.08025 18.54261 18.04121 17.57431 17.34561 17.16638
#> 8  14.79428 14.82241 14.86256 14.90004 14.93498 14.96751 15.00582 15.04283
#> 9  14.79428 14.82241 14.86256 14.90004 14.93498 14.96751 15.00582 15.04283
#> 10 14.79428 14.60434 14.39194 14.19374 14.00890 13.83678 13.74565 13.67209
#> 11 14.79428 14.60434 14.39194 14.19374 14.00890 13.83678 13.74565 13.67209
#> 12 26.91821 27.28466 27.68497 28.05851 28.40687 28.73126 28.88003 28.99345
#> 13 26.91821 27.06658 27.21434 27.35221 27.48078 27.60053 27.61986 27.62271
#> 14 25.56271 25.36274 25.18045 25.01038 24.85177 24.70402 24.72609 24.77290
#> 15 14.79428 15.04048 15.33319 15.60633 15.86107 16.09824 16.26599 16.41358
#> 16 14.79428 15.04048 15.33319 15.60633 15.86107 16.09824 16.26599 16.41358
#> 17 14.79428 14.60434 14.39194 14.19374 14.00890 13.83678 13.74565 13.67209
#> 18 14.79428 15.04048 15.33319 15.60633 15.86107 16.09824 16.26599 16.41358
#> 19 26.91821 27.28466 27.68497 28.05851 28.40687 28.73126 28.88003 28.99345
#> 20 26.91821 27.06658 27.21434 27.35221 27.48078 27.60053 27.61986 27.62271
#> 21 16.14978 16.52625 16.89645 17.24187 17.56400 17.86402 17.89959 17.89264
#>      0.5270   0.4915   0.4584   0.4275   0.3987   0.3718   0.3467   0.3234
#> 1  21.35941 21.35939 21.35937 21.35937 21.35936 21.35936 21.35936 20.92936
#> 2  21.35941 21.35939 21.35937 21.35937 21.35936 21.35936 21.35936 20.92936
#> 3  21.31585 21.31597 21.31603 21.31605 21.31607 21.31608 21.31608 22.14317
#> 4  13.65413 13.65403 13.65397 13.65395 13.65393 13.65393 13.65392 13.47572
#> 5  23.38662 23.38642 23.38632 23.38627 23.38625 23.38624 23.38623 22.96034
#> 6  17.12264 17.12235 17.12221 17.12214 17.12210 17.12209 17.12208 17.97286
#> 7  17.12264 17.12235 17.12221 17.12214 17.12210 17.12209 17.12208 17.97286
#> 8  15.05187 15.05190 15.05191 15.05192 15.05192 15.05192 15.05192 14.86583
#> 9  15.05187 15.05190 15.05191 15.05192 15.05192 15.05192 15.05192 14.86583
#> 10 13.65413 13.65403 13.65397 13.65395 13.65393 13.65393 13.65392 13.47572
#> 11 13.65413 13.65403 13.65397 13.65395 13.65393 13.65393 13.65392 13.47572
#> 12 29.02113 29.02133 29.02143 29.02147 29.02150 29.02151 29.02151 29.59681
#> 13 27.62339 27.62346 27.62349 27.62350 27.62351 27.62351 27.62351 28.20670
#> 14 24.78436 24.78429 24.78426 24.78425 24.78424 24.78423 24.78423 24.35045
#> 15 16.44961 16.44977 16.44985 16.44989 16.44991 16.44992 16.44992 16.25594
#> 16 16.44961 16.44977 16.44985 16.44989 16.44991 16.44992 16.44992 16.25594
#> 17 13.65413 13.65403 13.65397 13.65395 13.65393 13.65393 13.65392 13.47572
#> 18 16.44961 16.44977 16.44985 16.44989 16.44991 16.44992 16.44992 16.25594
#> 19 29.02113 29.02133 29.02143 29.02147 29.02150 29.02151 29.02151 29.59681
#> 20 27.62339 27.62346 27.62349 27.62350 27.62351 27.62351 27.62351 25.91685
#> 21 17.89091 17.89106 17.89114 17.89117 17.89119 17.89120 17.89121 18.72208
#>      0.3016   0.2812   0.2623   0.2446   0.2281   0.2128   0.1984   0.1850
#> 1  20.92930 20.92925 20.92921 21.69799 21.69764 21.69737 21.69716 21.69699
#> 2  20.92930 20.92925 20.92921 21.69799 21.69764 21.69737 21.69716 21.69699
#> 3  22.14336 22.14351 22.14363 22.19969 22.19993 22.20008 22.20018 22.20022
#> 4  13.47565 13.47560 13.47555 12.75254 12.75266 12.75278 12.75288 12.75298
#> 5  22.96002 22.95977 22.95957 22.60084 22.60052 22.60034 22.60025 22.60023
#> 6  17.97278 17.97272 17.97268 18.46362 18.46322 18.46289 18.46262 18.46241
#> 7  17.97278 17.97272 17.97268 18.46362 18.46322 18.46289 18.46262 18.46241
#> 8  14.86585 14.86586 14.86587 15.15173 15.15163 15.15154 15.15146 15.15140
#> 9  14.86585 14.86586 14.86587 15.15173 15.15163 15.15154 15.15146 15.15140
#> 10 13.47565 13.47560 13.47555 13.32312 13.32302 13.32296 13.32291 13.32288
#> 11 13.47565 13.47560 13.47555 13.03783 13.03784 13.03787 13.03790 13.03793
#> 12 29.59701 29.59717 29.59729 29.41439 29.41481 29.41513 29.41535 29.41551
#> 13 28.20682 28.20690 28.20697 28.51772 28.51780 28.51784 28.51786 28.51785
#> 14 24.35021 24.35003 24.34989 23.72574 23.72569 23.72570 23.72576 23.72585
#> 15 16.25604 16.25612 16.25619 16.88525 16.88518 16.88510 16.88501 16.88492
#> 16 16.25604 16.25612 16.25619 16.88525 16.88518 16.88510 16.88501 16.88492
#> 17 13.47565 13.47560 13.47555 12.75254 12.75266 12.75278 12.75288 12.75298
#> 18 16.25604 16.25612 16.25619 16.40977 16.40988 16.40994 16.40998 16.41000
#> 19 29.59701 29.59717 29.59729 29.41439 29.41481 29.41513 29.41535 29.41551
#> 20 25.91654 25.91630 25.91611 25.66404 25.66384 25.66377 25.66377 25.66383
#> 21 18.72245 18.72273 18.72295 17.49023 17.49118 17.49190 17.49242 17.49279
#>      0.1726   0.1609   0.1501   0.1400   0.1305   0.1217   0.1135   0.1059
#> 1  21.69686 21.25478 21.40268 21.40267 21.40268 21.40269 21.40272 21.38750
#> 2  21.69686 21.25478 21.40268 21.40267 21.40268 21.40269 21.40272 21.53468
#> 3  22.20024 21.74270 21.66420 21.66407 21.66398 21.66392 21.66388 21.48312
#> 4  12.75306 12.26686 12.46208 12.46188 12.46173 12.46161 12.46151 12.35953
#> 5  22.60025 22.47562 22.05058 22.05096 22.05125 22.05147 22.05163 22.75002
#> 6  18.46224 18.74670 18.45305 18.45335 18.45359 18.45379 18.45395 18.10001
#> 7  18.46224 18.74670 18.45305 18.45335 18.45359 18.45379 18.45395 18.25770
#> 8  15.15134 15.18851 15.22181 15.22189 15.22196 15.22202 15.22207 15.30199
#> 9  15.15134 15.18851 15.22181 15.22189 15.22196 15.22202 15.22207 15.40712
#> 10 13.32287 12.61164 12.67472 12.67465 12.67460 12.67458 12.67457 12.95865
#> 11 13.03796 12.71060 12.91372 12.91360 12.91351 12.91344 12.91339 13.17449
#> 12 29.41562 28.97494 29.05643 29.05606 29.05578 29.05557 29.05540 29.17459
#> 13 28.51784 29.70289 30.34575 30.34561 30.34548 30.34536 30.34525 30.29315
#> 14 23.72595 23.43551 23.72051 23.72054 23.72055 23.72054 23.72053 23.71663
#> 15 16.88484 16.80704 16.58701 16.58716 16.58730 16.58742 16.58751 16.30186
#> 16 16.88484 17.65364 17.66439 17.66461 17.66478 17.66490 17.66499 17.53912
#> 17 12.75306 13.39566 13.89860 13.89848 13.89836 13.89825 13.89815 13.74551
#> 18 16.41000 16.83449 16.81037 16.81045 16.81050 16.81054 16.81057 16.70155
#> 19 29.41562 28.97494 29.05643 29.05606 29.05578 29.05557 29.05540 29.02478
#> 20 25.66393 25.89402 25.32478 25.32516 25.32544 25.32564 25.32578 25.20951
#> 21 17.49305 17.83948 17.31535 17.31488 17.31450 17.31420 17.31396 17.27848
#>      0.0988   0.0921   0.0859   0.0801   0.0747   0.0697   0.0650   0.0606
#> 1  21.38726 21.38706 21.38691 21.38680 21.38671 21.38664 21.38659 21.38655
#> 2  21.53447 21.53430 21.53417 21.53406 21.53398 21.53392 21.53387 21.53384
#> 3  21.48286 21.48264 21.48247 21.48233 21.48221 21.48213 21.48206 21.48201
#> 4  12.35943 12.35934 12.35925 12.35917 12.35911 12.35905 12.35900 12.35895
#> 5  22.75043 22.75077 22.75106 22.75129 22.75149 22.75165 22.75177 22.75187
#> 6  18.09990 18.09985 18.09983 18.09984 18.09986 18.09990 18.09994 18.09998
#> 7  18.25763 18.25760 18.25760 18.25763 18.25766 18.25770 18.25774 18.25778
#> 8  15.30198 15.30198 15.30198 15.30199 15.30200 15.30201 15.30202 15.30203
#> 9  15.40713 15.40715 15.40717 15.40718 15.40720 15.40721 15.40722 15.40723
#> 10 12.95850 12.95838 12.95827 12.95818 12.95811 12.95805 12.95800 12.95796
#> 11 13.17449 13.17447 13.17445 13.17442 13.17439 13.17436 13.17433 13.17431
#> 12 29.17442 29.17426 29.17411 29.17397 29.17384 29.17373 29.17363 29.17355
#> 13 30.29337 30.29353 30.29364 30.29372 30.29377 30.29380 30.29382 30.29383
#> 14 23.71681 23.71696 23.71707 23.71716 23.71723 23.71728 23.71732 23.71735
#> 15 16.30170 16.30159 16.30153 16.30149 16.30147 16.30146 16.30147 16.30148
#> 16 17.53925 17.53937 17.53949 17.53959 17.53968 17.53976 17.53983 17.53988
#> 17 13.74574 13.74591 13.74603 13.74612 13.74619 13.74623 13.74625 13.74627
#> 18 16.70164 16.70172 16.70179 16.70184 16.70189 16.70194 16.70197 16.70200
#> 19 29.02458 29.02439 29.02422 29.02407 29.02393 29.02382 29.02372 29.02363
#> 20 25.20992 25.21027 25.21057 25.21083 25.21105 25.21123 25.21137 25.21149
#> 21 17.27849 17.27845 17.27839 17.27832 17.27823 17.27815 17.27807 17.27800
#>      0.0565   0.0527   0.0491   0.0458   0.0427   0.0399   0.0372   0.0347
#> 1  21.38653 21.38651 21.38650 21.38649 21.38649 21.38649 21.38649 21.38649
#> 2  21.53381 21.53380 21.53378 21.53377 21.53377 21.53377 21.53377 21.53377
#> 3  21.48197 21.48194 21.48191 21.48190 21.48189 21.48188 21.48188 21.48188
#> 4  12.35892 12.35889 12.35886 12.35884 12.35883 12.35881 12.35880 12.35880
#> 5  22.75195 22.75201 22.75206 22.75210 22.75213 22.75215 22.75216 22.75217
#> 6  18.10002 18.10005 18.10009 18.10012 18.10014 18.10017 18.10019 18.10020
#> 7  18.25782 18.25786 18.25789 18.25792 18.25795 18.25797 18.25799 18.25800
#> 8  15.30204 15.30205 15.30206 15.30206 15.30207 15.30207 15.30208 15.30208
#> 9  15.40724 15.40725 15.40726 15.40727 15.40727 15.40727 15.40728 15.40728
#> 10 12.95793 12.95791 12.95789 12.95787 12.95786 12.95785 12.95785 12.95785
#> 11 13.17429 13.17427 13.17425 13.17423 13.17422 13.17421 13.17420 13.17419
#> 12 29.17348 29.17342 29.17337 29.17333 29.17330 29.17327 29.17325 29.17324
#> 13 30.29383 30.29383 30.29382 30.29381 30.29380 30.29379 30.29378 30.29377
#> 14 23.71737 23.71739 23.71740 23.71740 23.71741 23.71741 23.71741 23.71741
#> 15 16.30149 16.30151 16.30153 16.30154 16.30156 16.30157 16.30159 16.30160
#> 16 17.53993 17.53997 17.54000 17.54002 17.54005 17.54006 17.54007 17.54008
#> 17 13.74627 13.74628 13.74627 13.74627 13.74626 13.74625 13.74625 13.74624
#> 18 16.70202 16.70204 16.70205 16.70207 16.70207 16.70208 16.70209 16.70209
#> 19 29.02356 29.02350 29.02346 29.02342 29.02339 29.02336 29.02334 29.02333
#> 20 25.21159 25.21167 25.21173 25.21178 25.21181 25.21184 25.21186 25.21188
#> 21 17.27793 17.27788 17.27782 17.27778 17.27774 17.27771 17.27769 17.27767
#>      0.0323   0.0302   0.0281   0.0262   0.0245   0.0228   0.0213   0.0198
#> 1  21.38649 21.38649 21.38649 21.38649 21.38650 21.38650 21.38654 21.40840
#> 2  21.53377 21.53377 21.53377 21.53377 21.53377 21.53377 21.53380 21.55796
#> 3  21.48187 21.48187 21.48188 21.48188 21.48188 21.48188 21.48184 21.42715
#> 4  12.35879 12.35878 12.35878 12.35878 12.35878 12.35878 12.35892 12.44139
#> 5  22.75218 22.75218 22.75219 22.75219 22.75219 22.75219 22.75219 22.76141
#> 6  18.10022 18.10023 18.10024 18.10024 18.10025 18.10025 18.10046 18.09522
#> 7  18.25801 18.25802 18.25803 18.25804 18.25804 18.25805 18.25824 18.25547
#> 8  15.30208 15.30209 15.30209 15.30209 15.30209 15.30209 15.30226 15.40356
#> 9  15.40728 15.40728 15.40729 15.40729 15.40729 15.40729 15.40745 15.51039
#> 10 12.95784 12.95784 12.95784 12.95784 12.95784 12.95784 12.95772 12.82537
#> 11 13.17419 13.17418 13.17418 13.17418 13.17417 13.17417 13.17407 13.08627
#> 12 29.17323 29.17322 29.17321 29.17320 29.17320 29.17320 29.17309 29.20679
#> 13 30.29376 30.29376 30.29375 30.29375 30.29374 30.29374 30.29349 30.19462
#> 14 23.71741 23.71741 23.71741 23.71740 23.71740 23.71740 23.71749 23.80532
#> 15 16.30161 16.30161 16.30162 16.30163 16.30163 16.30163 16.30169 16.29621
#> 16 17.54009 17.54010 17.54010 17.54010 17.54011 17.54011 17.54010 17.51865
#> 17 13.74623 13.74623 13.74622 13.74622 13.74621 13.74621 13.74626 13.78833
#> 18 16.70209 16.70209 16.70210 16.70210 16.70210 16.70210 16.70193 16.58164
#> 19 29.02332 29.02331 29.02330 29.02330 29.02330 29.02329 29.02320 29.05403
#> 20 25.21189 25.21190 25.21190 25.21190 25.21191 25.21191 25.21180 25.19959
#> 21 17.27765 17.27764 17.27763 17.27762 17.27761 17.27761 17.27747 17.28224
#>      0.0185   0.0173   0.0161   0.0150   0.0140   0.0131   0.0122   0.0114
#> 1  21.40827 21.40815 21.33542 21.33529 21.33517 21.33505 21.33494 21.33484
#> 2  21.55785 21.55775 21.53136 21.53127 21.53119 21.53112 21.53105 21.53098
#> 3  21.42676 21.42644 21.37365 21.37354 21.37343 21.37334 21.37325 21.37317
#> 4  12.44141 12.44142 12.30384 12.30352 12.30322 12.30294 12.30267 12.30242
#> 5  22.76180 22.76215 22.65828 22.65811 22.65795 22.65780 22.65766 22.65752
#> 6  18.09527 18.09533 18.14865 18.14888 18.14910 18.14930 18.14950 18.14968
#> 7  18.25554 18.25561 18.26615 18.26628 18.26640 18.26652 18.26663 18.26673
#> 8  15.40375 15.40391 15.44158 15.44160 15.44161 15.44162 15.44163 15.44164
#> 9  15.51060 15.51077 15.53683 15.53681 15.53678 15.53676 15.53673 15.53671
#> 10 12.82494 12.82458 12.88201 12.88200 12.88200 12.88200 12.88199 12.88199
#> 11 13.08604 13.08584 13.22823 13.22837 13.22850 13.22863 13.22875 13.22885
#> 12 29.20658 29.20639 29.30158 29.30162 29.30166 29.30170 29.30173 29.30176
#> 13 30.19460 30.19459 30.11271 30.11268 30.11265 30.11263 30.11261 30.11259
#> 14 23.80563 23.80589 23.83717 23.83725 23.83732 23.83738 23.83744 23.83750
#> 15 16.29612 16.29604 16.27115 16.27116 16.27118 16.27119 16.27121 16.27122
#> 16 17.51878 17.51890 17.50279 17.50286 17.50292 17.50298 17.50303 17.50309
#> 17 13.78860 13.78882 13.84301 13.84311 13.84320 13.84329 13.84337 13.84344
#> 18 16.58149 16.58139 16.50507 16.50501 16.50495 16.50490 16.50486 16.50481
#> 19 29.05381 29.05360 29.09905 29.09904 29.09903 29.09902 29.09901 29.09900
#> 20 25.20004 25.20044 25.29089 25.29124 25.29157 25.29188 25.29218 25.29245
#> 21 17.28212 17.28200 17.23058 17.23036 17.23015 17.22996 17.22977 17.22960
#>      0.0106   0.0099   0.0092   0.0086   0.0080   0.0075   0.0070   0.0065
#> 1  21.33474 21.33465 21.33457 21.33449 21.33441 21.33434 21.33428 21.33421
#> 2  21.53092 21.53086 21.53080 21.53075 21.53070 21.53065 21.53061 21.53057
#> 3  21.37309 21.37302 21.37295 21.37288 21.37283 21.37277 21.37272 21.37267
#> 4  12.30218 12.30195 12.30174 12.30153 12.30135 12.30117 12.30100 12.30084
#> 5  22.65739 22.65727 22.65716 22.65705 22.65695 22.65686 22.65677 22.65669
#> 6  18.14986 18.15002 18.15018 18.15033 18.15047 18.15060 18.15072 18.15084
#> 7  18.26683 18.26693 18.26702 18.26710 18.26718 18.26726 18.26733 18.26740
#> 8  15.44164 15.44165 15.44165 15.44165 15.44165 15.44165 15.44165 15.44165
#> 9  15.53669 15.53666 15.53664 15.53662 15.53659 15.53657 15.53655 15.53653
#> 10 12.88199 12.88199 12.88199 12.88199 12.88199 12.88199 12.88199 12.88199
#> 11 13.22896 13.22905 13.22914 13.22923 13.22931 13.22938 13.22945 13.22952
#> 12 29.30179 29.30181 29.30183 29.30185 29.30187 29.30189 29.30190 29.30191
#> 13 30.11257 30.11256 30.11255 30.11254 30.11253 30.11252 30.11251 30.11251
#> 14 23.83756 23.83760 23.83765 23.83769 23.83773 23.83777 23.83780 23.83783
#> 15 16.27123 16.27125 16.27126 16.27127 16.27129 16.27130 16.27131 16.27133
#> 16 17.50314 17.50319 17.50323 17.50328 17.50332 17.50336 17.50340 17.50344
#> 17 13.84351 13.84358 13.84364 13.84369 13.84375 13.84379 13.84384 13.84388
#> 18 16.50478 16.50474 16.50471 16.50468 16.50466 16.50463 16.50461 16.50459
#> 19 29.09899 29.09898 29.09897 29.09896 29.09895 29.09894 29.09893 29.09892
#> 20 25.29272 25.29296 25.29320 25.29342 25.29363 25.29382 25.29401 25.29418
#> 21 17.22943 17.22928 17.22913 17.22899 17.22885 17.22873 17.22861 17.22850
#>      0.0061   0.0057   0.0053   0.0049
#> 1  21.33416 21.33410 21.33405 21.33400
#> 2  21.53053 21.53049 21.53045 21.53042
#> 3  21.37263 21.37259 21.37255 21.37252
#> 4  12.30069 12.30055 12.30042 12.30030
#> 5  22.65661 22.65654 22.65647 22.65641
#> 6  18.15095 18.15106 18.15116 18.15125
#> 7  18.26747 18.26753 18.26759 18.26764
#> 8  15.44165 15.44165 15.44164 15.44164
#> 9  15.53651 15.53649 15.53647 15.53645
#> 10 12.88199 12.88199 12.88199 12.88199
#> 11 13.22958 13.22963 13.22969 13.22974
#> 12 29.30192 29.30193 29.30194 29.30194
#> 13 30.11250 30.11250 30.11250 30.11250
#> 14 23.83786 23.83789 23.83791 23.83793
#> 15 16.27134 16.27135 16.27136 16.27137
#> 16 17.50347 17.50350 17.50354 17.50357
#> 17 13.84392 13.84395 13.84399 13.84402
#> 18 16.50458 16.50456 16.50455 16.50454
#> 19 29.09891 29.09890 29.09889 29.09888
#> 20 25.29434 25.29450 25.29464 25.29478
#> 21 17.22839 17.22829 17.22820 17.22811
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 21.4 14.3 22.8 19.2 17.8 17.3 15.2 10.4 14.7 32.4 30.4 21.5 15.5
#> [16] 15.2 13.3 19.2 27.3 26.0 15.8
#> 
#> $X
#>               am       carb        cyl       disp       drat       gear
#>  [1,]  1.4142136  1.0391690 -0.2236068 -0.6708843  0.4743760  0.6495191
#>  [2,]  1.4142136  1.0391690 -0.2236068 -0.6708843  0.4743760  0.6495191
#>  [3,] -0.7071068 -1.4788174 -0.2236068  0.1401354 -1.0054708 -0.8660254
#>  [4,] -0.7071068  1.0391690  0.9503289  0.9842578 -0.7708610 -0.8660254
#>  [5,] -0.7071068 -0.6394886 -1.3975425 -0.8297779  0.5104698  0.6495191
#>  [6,] -0.7071068  1.0391690 -0.2236068 -0.6079889  0.5104698  0.6495191
#>  [7,] -0.7071068  1.0391690 -0.2236068 -0.6079889  0.5104698  0.6495191
#>  [8,] -0.7071068  0.1998402  0.9503289  0.2874430 -1.0235177 -0.8660254
#>  [9,] -0.7071068  0.1998402  0.9503289  0.2874430 -1.0235177 -0.8660254
#> [10,] -0.7071068  1.0391690  0.9503289  1.8118289 -1.1498461 -0.8660254
#> [11,] -0.7071068  1.0391690  0.9503289  1.6463146 -0.7347671 -0.8660254
#> [12,]  1.4142136 -1.4788174 -1.3975425 -1.3436995  0.7992204  0.6495191
#> [13,]  1.4142136 -0.6394886 -1.3975425 -1.3685266  2.3332079  0.6495191
#> [14,] -0.7071068 -1.4788174 -1.3975425 -1.0010851  0.1134377 -0.8660254
#> [15,] -0.7071068 -0.6394886  0.9503289  0.6366780 -1.5829720 -0.8660254
#> [16,] -0.7071068 -0.6394886  0.9503289  0.5208180 -0.8791424 -0.8660254
#> [17,] -0.7071068  1.0391690  0.9503289  0.9015007  0.1675785 -0.8660254
#> [18,] -0.7071068 -0.6394886  0.9503289  1.3152862 -1.0054708 -0.8660254
#> [19,]  1.4142136 -1.4788174 -1.3975425 -1.3412168  0.7992204  0.6495191
#> [20,]  1.4142136 -0.6394886 -1.3975425 -0.9994300  1.4308623  2.1650635
#> [21,]  1.4142136  1.0391690  0.9503289  0.9097764  1.0518772  2.1650635
#>                hp         qsec         vs          wt
#>  [1,] -0.58050183 -0.775716330 -0.7844645 -0.69666175
#>  [2,] -0.58050183 -0.458635202 -0.7844645 -0.42502758
#>  [3,] -0.58050183  0.911608242  1.2747549 -0.06284869
#>  [4,]  1.56516008 -1.126770436 -0.7844645  0.31530868
#>  [5,] -0.81890871  2.870716638  1.2747549 -0.13208877
#>  [6,] -0.37388253  0.266121661  1.2747549  0.17682852
#>  [7,] -0.37388253  0.605851440  1.2747549  0.17682852
#>  [8,]  0.53206361 -0.130229749 -0.7844645  0.48574581
#>  [9,]  0.53206361  0.096256771 -0.7844645  0.53900741
#> [10,]  1.08834632 -0.005662163 -0.7844645  2.29024886
#> [11,]  1.32675320 -0.232148683 -0.7844645  2.20609553
#> [12,] -1.27982867  0.928594731  1.2747549 -1.14405920
#> [13,] -1.50234176  0.390689247  1.2747549 -1.76721994
#> [14,] -0.78712112  1.234351533  1.2747549 -0.86177271
#> [15,]  0.05524985 -0.543567647 -0.7844645  0.26204708
#> [16,]  0.05524985 -0.300094639 -0.7844645  0.17150236
#> [17,]  1.56516008 -1.370243444 -0.7844645  0.60292133
#> [18,]  0.45259465 -0.441648714 -0.7844645  0.60824749
#> [19,] -1.27982867  0.605851440  1.2747549 -1.42634569
#> [20,] -0.88248388 -0.639824418 -0.7844645 -1.20797312
#> [21,]  1.86714213 -1.885500277 -0.7844645 -0.11078413
#> attr(,"center")
#>  [1]   0.3333333   2.7619048   6.3809524 241.0666667   3.6371429   3.5714286
#>  [7] 146.5238095  17.8300000   0.3809524   3.2740000
#> attr(,"scale")
#>  [1]   0.4714045   1.1914282   1.7036708 120.8355487   0.5541114   0.6598289
#>  [7]  62.9176476   1.7661095   0.4856209   0.9387626
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 34
#> 
#> $lambda.min
#> [1] 0.4914824
#> 
#> $null.dev
#> [1] 31.28522
#> 
#> $Bias
#> [1] 3.276872
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 13.30636