Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 35.79277 34.52659 32.79511 30.07109 27.03206 24.58576 22.64171 21.12212
#>   [9] 19.96017 19.09850 18.48797 18.08651 17.85814 17.77219 17.80251 17.58167
#>  [17] 16.95920 16.45216 16.07448 16.04262 16.07979 16.11793 16.17849 16.31560
#>  [25] 17.10374 18.19645 18.42799 17.86369 17.55825 17.60332 16.08316 16.02563
#>  [33] 15.95050 15.88965 15.86091 18.89061 18.48008 17.95954 17.00973 16.78807
#>  [41] 16.78797 16.78791 16.88301 15.90086 15.88299 17.85805 17.85837 18.64348
#>  [49] 18.76961 21.25444 22.05770 24.79038 24.79880 31.97750 31.65432 31.65588
#>  [57] 33.76145 33.75926 33.82933 34.69421 34.69437 35.33384 35.33457 35.50895
#>  [65] 35.85247 35.85319 35.85389 35.85458 35.86242 35.89215 35.89269 35.88394
#>  [73] 35.86590 35.52728 35.51971 35.48468 37.90322 37.90413 37.90504 37.90596
#>  [81] 37.90687 37.90779 37.90871 37.90971 37.91328 37.91419 37.91711 38.04944
#>  [89] 38.03582 38.03687 38.03789 38.04100 38.04210 38.01086 38.01210 38.01333
#>  [97] 38.01454 38.01573 38.01690 38.01806
#> 
#> $cvse
#>   [1] 13.111227 13.186907 13.045674 12.181033 10.975326  9.889620  8.916685
#>   [8]  8.053043  7.298404  6.654970  6.126409  5.716300  5.426081  5.252975
#>  [15]  5.188769  5.094858  4.871058  4.672638  4.500785  4.458428  4.433742
#>  [22]  4.412579  4.396379  4.381016  4.276142  4.697417  4.681974  4.599030
#>  [29]  4.680242  4.687101  4.693398  4.686531  4.676816  4.666971  4.651736
#>  [36]  5.299700  5.349105  5.220989  4.679474  4.547818  4.547811  4.547810
#>  [43]  4.378498  4.135176  4.140494  4.523586  4.523688  4.475375  4.429534
#>  [50]  5.073510  5.114034  7.228878  7.228255 11.070949 11.036657 11.037695
#>  [57] 10.998608 11.000129 10.997013 11.021909 11.022018 11.597279 11.597605
#>  [64] 11.587976 11.557333 11.557616 11.557901 11.558188 11.557868 11.554472
#>  [71] 11.554772 11.554937 11.554933 11.502751 11.502901 11.505510 11.740926
#>  [78] 11.741381 11.741832 11.742278 11.742721 11.743159 11.743592 11.744013
#>  [85] 11.744340 11.744761 11.744984 11.732570 11.734988 11.735373 11.735756
#>  [92] 11.730687 11.731045 11.735634 11.735960 11.736283 11.736601 11.736916
#>  [99] 11.737227 11.737534
#> 
#> $fold
#>  [1]  3  7  9  7  1  1  2  9  4  8 10  8  6 10  2  4  1  5  5  3  6
#> 
#> $lambda
#>   [1] 4.989134836 4.652883847 4.339295048 4.046841085 3.774097540 3.519735997
#>   [7] 3.282517571 3.061286873 2.854966384 2.662551205 2.483104165 2.315751255
#>  [13] 2.159677371 2.014122344 1.878377239 1.751780900 1.633716730 1.523609690
#>  [19] 1.420923497 1.325158009 1.235846794 1.152554856 1.074876516 1.002433437
#>  [25] 0.934872778 0.871865482 0.813104666 0.758304133 0.707196973 0.659534264
#>  [31] 0.615083862 0.573629268 0.534968575 0.498913484 0.465288385 0.433929505
#>  [37] 0.404684108 0.377409754 0.351973600 0.328251757 0.306128687 0.285496638
#>  [43] 0.266255120 0.248310416 0.231575125 0.215967737 0.201412234 0.187837724
#>  [49] 0.175178090 0.163371673 0.152360969 0.142092350 0.132515801 0.123584679
#>  [55] 0.115255486 0.107487652 0.100243344 0.093487278 0.087186548 0.081310467
#>  [61] 0.075830413 0.070719697 0.065953426 0.061508386 0.057362927 0.053496858
#>  [67] 0.049891348 0.046528838 0.043392950 0.040468411 0.037740975 0.035197360
#>  [73] 0.032825176 0.030612869 0.028549664 0.026625512 0.024831042 0.023157513
#>  [79] 0.021596774 0.020141223 0.018783772 0.017517809 0.016337167 0.015236097
#>  [85] 0.014209235 0.013251580 0.012358468 0.011525549 0.010748765 0.010024334
#>  [91] 0.009348728 0.008718655 0.008131047 0.007583041 0.007071970 0.006595343
#>  [97] 0.006150839 0.005736293 0.005349686 0.004989135
#> 
#> $fit
#> $beta
#>               4.9891     4.6529     4.3393    4.0468    3.7741    3.5197
#> (Intercept) 19.75238 21.7664054 23.6446914 25.396387 27.030025 28.553560
#> am           0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> cyl          0.00000 -0.3109891 -0.6010185 -0.871501 -1.123754 -1.359006
#> disp         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> wt           0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#>                3.2825    3.0613    2.8550    2.6626    2.4831    2.3158
#> (Intercept) 29.974415 31.299508 32.535295 33.687794 34.762618 35.765003
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.578402 -1.783012 -1.973832 -2.151792 -2.317757 -2.472537
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                2.1597    2.0141    1.8784    1.7518    1.6337    1.5236
#> (Intercept) 36.699830 37.571653 38.384718 39.142986 39.674483 39.674483
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.616886 -2.751505 -2.877052 -2.994138 -3.076207 -3.076207
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.4209    1.3252    1.2358    1.1526    1.0749    1.0024
#> (Intercept) 39.674483 39.674483 39.674483 39.674483 39.674483 39.674483
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -3.076207 -3.076207 -3.076207 -3.076207 -3.076207 -3.076207
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                0.9349    0.8719    0.8131    0.7583    0.7072      0.6595
#> (Intercept) 39.674483 40.342275 40.341868 40.341600 40.341425 40.31759224
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000 -0.04274529
#> cyl         -3.076207 -1.726880 -1.726666 -1.726525 -1.726433 -1.69842371
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> wt           0.000000 -2.899852 -2.900154 -2.900352 -2.900482 -2.91015011
#>                 0.6151     0.5736    0.5350     0.4989    0.4653     0.4339
#> (Intercept) 40.2842605 40.2531536 40.224500 40.1974377 40.172182 40.1486285
#> am           0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.0000000
#> carb        -0.1014671 -0.1562375 -0.207184 -0.2548234 -0.299258 -0.3406982
#> cyl         -1.6596347 -1.6234463 -1.589933 -1.5584529 -1.529084 -1.5016938
#> disp         0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.0000000
#> drat         0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.0000000
#> gear         0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.0000000
#> hp           0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.0000000
#> qsec         0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.0000000
#> vs           0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.0000000
#> wt          -2.9238702 -2.9366804 -2.948388 -2.9595341 -2.969941 -2.9796464
#>                0.4047    0.3774    0.3520    0.3283    0.3061    0.2855
#> (Intercept) 34.090720 34.092009 34.092998 34.093745 34.094299 34.094702
#> am           3.767437  3.766713  3.766123  3.765652  3.765283  3.764999
#> carb        -1.121432 -1.121258 -1.121124 -1.121023 -1.120948 -1.120894
#> cyl         -1.110023 -1.110173 -1.110260 -1.110304 -1.110322 -1.110322
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -1.681201 -1.681363 -1.681539 -1.681709 -1.681865 -1.682001
#>                0.2663    0.2483     0.2316     0.2160     0.2014     0.1878
#> (Intercept) 34.094990 34.095192  9.6519448  9.6502942  9.6491570  9.6483681
#> am           3.764785  3.764625  5.1385800  5.1392295  5.1396918  5.1400223
#> carb        -1.120855 -1.120827 -0.7618125 -0.7618893 -0.7619463 -0.7619886
#> cyl         -1.110312 -1.110298  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.000000  0.000000  1.0688399  1.0688771  1.0689017  1.0689180
#> vs           0.000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -1.682117 -1.682212 -2.6665339 -2.6662428 -2.6660345 -2.6658851
#>                 0.1752     0.1634     0.1524     0.1421     0.1325     0.1236
#> (Intercept)  9.6478172  9.6474302  9.6471569  9.6469629  8.9662366  7.8438882
#> am           5.1402592  5.1404296  5.1405525  5.1406413  5.0658612  4.9373324
#> carb        -0.7620198 -0.7620428 -0.7620598 -0.7620723 -0.7666852 -0.7741948
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.1783336  0.4756249
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         1.0689290  1.0689364  1.0689414  1.0689449  1.0687590  1.0680966
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.6657775 -2.6656999 -2.6656438 -2.6656032 -2.6343103 -2.5837268
#>                 0.1153     0.1075     0.1002      0.0935      0.0872
#> (Intercept)  6.7970305  6.6915798  6.6929041 -5.47720008 -5.48063050
#> am           4.8173059  4.8060507  4.8052101  4.86177743  4.86179276
#> carb        -0.7811925 -0.7819395 -0.7819297 -0.29874582 -0.29858114
#> cyl          0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> disp         0.0000000  0.0000000  0.0000000  0.01563002  0.01563469
#> drat         0.7530018  0.7803784  0.7807008  1.06240546  1.06241107
#> gear         0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> hp           0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> qsec         1.0674707  1.0674758  1.0673966  1.73919889  1.73940693
#> vs           0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> wt          -2.5365776 -2.5316576 -2.5318844 -4.34980141 -4.35038072
#>                  0.0813      0.0758      0.0707      0.0660      0.0615
#> (Intercept) -5.21381750 -5.21717956 -5.21993351 -5.22217366  0.66678306
#> am           4.68730562  4.68746047  4.68757285  4.68765135  4.36387697
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000 -0.41151431
#> disp         0.02212072  0.02212039  0.02211991  0.02211935  0.02370609
#> drat         1.17362241  1.17350517  1.17341189  1.17333847  0.98679823
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.01128222 -0.01127544 -0.01126974 -0.01126498 -0.01094334
#> qsec         1.75483342  1.75502219  1.75517587  1.75530005  1.56469874
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -4.82552704 -4.82571073 -4.82584941 -4.82595168 -4.67596803
#>                  0.0574      0.0535       0.0499       0.0465       0.0434
#> (Intercept)  0.66523675  0.66331786 -0.830388900 -0.830740018 -0.831055605
#> am           4.36392658  4.36401822  4.168386723  4.167893053  4.167441038
#> carb         0.00000000  0.00000000  0.000000000  0.000000000  0.000000000
#> cyl         -0.41118058 -0.41086661 -0.642369469 -0.642689784 -0.642978536
#> disp         0.02370431  0.02370301  0.024548979  0.024547663  0.024546413
#> drat         0.98710277  0.98737592  1.039780951  1.039900445  1.040012132
#> gear         0.00000000  0.00000000  0.000000000  0.000000000  0.000000000
#> hp          -0.01094985 -0.01095539 -0.007303022 -0.007298189 -0.007293829
#> qsec         1.56466157  1.56465319  1.743520385  1.743636067  1.743739676
#> vs           0.00000000  0.00000000 -0.933434783 -0.934318768 -0.935118446
#> wt          -4.67584382 -4.67577545 -4.885088944 -4.885064773 -4.885038008
#>                   0.0405       0.0377       0.0352      0.0328       0.0306
#> (Intercept) -0.831339614 -0.831595531 -5.019017583 -5.02386774 -5.028552403
#> am           4.167027745  4.166650393  3.559967197  3.55915571  3.558378697
#> carb         0.000000000  0.000000000  0.000000000  0.00000000  0.000000000
#> cyl         -0.643238369 -0.643471747 -0.730636455 -0.73091623 -0.731183431
#> disp         0.024545228  0.024544108  0.026251431  0.02625379  0.026256096
#> drat         1.040116372  1.040213528  1.074973217  1.07503864  1.075100891
#> gear         0.000000000  0.000000000  0.561511538  0.56205775  0.562583293
#> hp          -0.007289903 -0.007286372 -0.007695567 -0.00769264 -0.007689837
#> qsec         1.743832326  1.743915039  1.923768178  1.92405890  1.924339162
#> vs          -0.935840753 -0.936492144 -1.564406563 -1.56557567 -1.566696778
#> wt          -4.885009306 -4.884979238 -5.022109275 -5.02239517 -5.022673299
#>                   0.0285       0.0266       0.0248       0.0232       0.0216
#> (Intercept) -5.033075182 -5.037439774 -5.041649934 -5.045709475 -5.049622259
#> am           3.557634850  3.556922881  3.556241512  3.555589481  3.554965553
#> carb         0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> cyl         -0.731438806 -0.731683074 -0.731916903 -0.732140918 -0.732355699
#> disp         0.026258337  0.026260519  0.026262641  0.026264705  0.026266711
#> drat         1.075160055  1.075216217  1.075269473  1.075319926  1.075367683
#> gear         0.563088705  0.563574538  0.564041357  0.564489738  0.564920262
#> hp          -0.007687149 -0.007684567 -0.007682085 -0.007679695 -0.007677393
#> qsec         1.924609351  1.924869825  1.925120930  1.925362999  1.925596351
#> vs          -1.567772113 -1.568803800 -1.569793869 -1.570744251 -1.571656785
#> wt          -5.022943837 -5.023206920 -5.023462680 -5.023711233 -5.023952691
#>                   0.0201       0.0188       0.0175       0.0163      0.0152
#> (Intercept) -5.053392184 -5.057023183 -5.095367633 -5.093991893 -5.09263850
#> am           3.554368517  3.553797193  3.430330311  3.430286058  3.43024773
#> carb         0.000000000  0.000000000 -0.151653001 -0.151844091 -0.15202822
#> cyl         -0.732561787 -0.732759682 -0.757733160 -0.757612370 -0.75749399
#> disp         0.026268660  0.026270553  0.023979477  0.023974785  0.02397025
#> drat         1.075412856  1.075455559  1.055297200  1.055282053  1.05526723
#> gear         0.565333514  0.565730081  0.764255028  0.764436554  0.76460863
#> hp          -0.007675172 -0.007673027 -0.005665681 -0.005665729 -0.00566580
#> qsec         1.925821290  1.926038112  1.897182754  1.897017009  1.89685543
#> vs          -1.572533219 -1.573375212 -1.814345320 -1.814188842 -1.81403063
#> wt          -5.024187161 -5.024414743 -4.751443758 -4.750882915 -4.75034040
#>                   0.0142       0.0133       0.0124       0.0115       0.0107
#> (Intercept) -5.091307583 -5.089999256 -5.088713588 -5.087450618 -5.086210356
#> am           3.430214980  3.430187477  3.430164906  3.430146968  3.430133377
#> carb        -0.152205620 -0.152376531 -0.152541176 -0.152699772 -0.152852529
#> cyl         -0.757378014 -0.757264412 -0.757153167 -0.757044256 -0.756937655
#> disp         0.023965857  0.023961611  0.023957506  0.023953536  0.023949697
#> drat         1.055252735  1.055238556  1.055224693  1.055211140  1.055197894
#> gear         0.764771680  0.764926101  0.765072281  0.765210590  0.765341383
#> hp          -0.005665892 -0.005666004 -0.005666134 -0.005666281 -0.005666444
#> qsec         1.896697946  1.896544464  1.896394906  1.896249192  1.896107243
#> vs          -1.813870987 -1.813710201 -1.813548541 -1.813386259 -1.813223591
#> wt          -4.749815641 -4.749308101 -4.748817245 -4.748342559 -4.747883541
#>                   0.0100       0.0093       0.0087       0.0081       0.0076
#> (Intercept) -5.084992786 -5.083797865 -5.082625526 -5.081475680 -5.080348218
#> am           3.430123861  3.430118161  3.430116031  3.430117236  3.430121555
#> carb        -0.152999650 -0.153141332 -0.153277764 -0.153409131 -0.153535611
#> cyl         -0.756833341 -0.756731287 -0.756631466 -0.756533850 -0.756438410
#> disp         0.023945985  0.023942397  0.023938928  0.023935574  0.023932333
#> drat         1.055184950  1.055172305  1.055159952  1.055147889  1.055136109
#> gear         0.765465001  0.765581770  0.765692004  0.765796004  0.765894058
#> hp          -0.005666622 -0.005666812 -0.005667015 -0.005667229 -0.005667453
#> qsec         1.895968979  1.895834323  1.895703196  1.895575523  1.895451228
#> vs          -1.813060759 -1.812897969 -1.812735414 -1.812573274 -1.812411716
#> wt          -4.747439705 -4.747010577 -4.746595699 -4.746194626 -4.745806923
#>                   0.0071       0.0066       0.0062       0.0057       0.0053
#> (Intercept) -5.079243011 -5.078159910 -5.077098752 -5.076059358 -5.075041534
#> am           3.430128775  3.430138696  3.430151125  3.430165883  3.430182797
#> carb        -0.153657375 -0.153774589 -0.153887414 -0.153996006 -0.154100513
#> cyl         -0.756345117 -0.756253941 -0.756164851 -0.756077816 -0.755992805
#> disp         0.023929200  0.023926171  0.023923245  0.023920417  0.023917685
#> drat         1.055124610  1.055113385  1.055102431  1.055091742  1.055081314
#> gear         0.765986441  0.766073419  0.766155245  0.766232162  0.766304403
#> hp          -0.005667686 -0.005667927 -0.005668176 -0.005668431 -0.005668693
#> qsec         1.895330237  1.895212476  1.895097873  1.894986355  1.894877853
#> vs          -1.812250896 -1.812090958 -1.811932037 -1.811774254 -1.811617726
#> wt          -4.745432172 -4.745069965 -4.744719904 -4.744381607 -4.744054701
#>                   0.0050
#> (Intercept) -5.074045074
#> am           3.430201704
#> carb        -0.154201082
#> cyl         -0.755909785
#> disp         0.023915045
#> drat         1.055071142
#> gear         0.766372192
#> hp          -0.005668959
#> qsec         1.894772298
#> vs          -1.811462556
#> wt          -4.743738822
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1   1   1  52   1   1   1   9   9   9   8   8   8   8
#>  [37]  38   1   1   1   1   1   1   1  44   1   1   1   1   1   1   1  23  26
#>  [55]  26   6   1 125   1  81   1   1   1  54   1   1  67   1   1   1   1  91
#>  [73]   1   1   1   1   1   1   1   1   1 107   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.989134836 4.652883847 4.339295048 4.046841085 3.774097540 3.519735997
#>   [7] 3.282517571 3.061286873 2.854966384 2.662551205 2.483104165 2.315751255
#>  [13] 2.159677371 2.014122344 1.878377239 1.751780900 1.633716730 1.523609690
#>  [19] 1.420923497 1.325158009 1.235846794 1.152554856 1.074876516 1.002433437
#>  [25] 0.934872778 0.871865482 0.813104666 0.758304133 0.707196973 0.659534264
#>  [31] 0.615083862 0.573629268 0.534968575 0.498913484 0.465288385 0.433929505
#>  [37] 0.404684108 0.377409754 0.351973600 0.328251757 0.306128687 0.285496638
#>  [43] 0.266255120 0.248310416 0.231575125 0.215967737 0.201412234 0.187837724
#>  [49] 0.175178090 0.163371673 0.152360969 0.142092350 0.132515801 0.123584679
#>  [55] 0.115255486 0.107487652 0.100243344 0.093487278 0.087186548 0.081310467
#>  [61] 0.075830413 0.070719697 0.065953426 0.061508386 0.057362927 0.053496858
#>  [67] 0.049891348 0.046528838 0.043392950 0.040468411 0.037740975 0.035197360
#>  [73] 0.032825176 0.030612869 0.028549664 0.026625512 0.024831042 0.023157513
#>  [79] 0.021596774 0.020141223 0.018783772 0.017517809 0.016337167 0.015236097
#>  [85] 0.014209235 0.013251580 0.012358468 0.011525549 0.010748765 0.010024334
#>  [91] 0.009348728 0.008718655 0.008131047 0.007583041 0.007071970 0.006595343
#>  [97] 0.006150839 0.005736293 0.005349686 0.004989135
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 25
#> 
#> $loss
#>   [1] 739.9124 639.5658 555.6110 485.6890 427.7632 380.0766 341.1138 309.5691
#>   [9] 284.3186 264.3949 248.9671 237.3213 228.8455 223.0154 219.3825 217.5636
#>  [17] 217.1916 217.1916 217.1916 217.1916 217.1916 217.1916 217.1916 217.1916
#>  [25] 217.1916 164.4378 164.4378 164.4378 164.4378 162.2808 159.6261 157.4721
#>  [33] 155.7474 154.3779 153.3125 152.5032 128.9273 128.9273 128.9273 128.9273
#>  [41] 128.9273 128.9273 128.9273 128.9273 121.6772 121.6772 121.6772 121.6772
#>  [49] 121.6772 121.6772 121.6772 121.6772 121.2734 120.8318 120.6807 120.6795
#>  [57] 120.6795 116.8294 116.8294 115.5112 115.5112 115.5112 115.5112 114.6295
#>  [65] 114.6295 114.6295 114.1437 114.1437 114.1436 114.1436 114.1436 113.7925
#>  [73] 113.7924 113.7924 113.7924 113.7924 113.7923 113.7923 113.7923 113.7923
#>  [81] 113.7923 113.6834 113.6833 113.6833 113.6833 113.6833 113.6833 113.6833
#>  [89] 113.6832 113.6832 113.6832 113.6832 113.6832 113.6832 113.6832 113.6832
#>  [97] 113.6832 113.6832 113.6832 113.6832
#> 
#> $linear.predictors
#>      4.9891   4.6529   4.3393   4.0468   3.7741   3.5197   3.2825   3.0613
#> 1  19.75238 19.90047 20.03858 20.16738 20.28750 20.39953 20.50400 20.60143
#> 2  19.75238 19.90047 20.03858 20.16738 20.28750 20.39953 20.50400 20.60143
#> 3  19.75238 20.52245 21.24062 21.91038 22.53501 23.11754 23.66081 24.16746
#> 4  19.75238 19.90047 20.03858 20.16738 20.28750 20.39953 20.50400 20.60143
#> 5  19.75238 19.27849 18.83654 18.42438 18.03999 17.68152 17.34720 17.03541
#> 6  19.75238 19.90047 20.03858 20.16738 20.28750 20.39953 20.50400 20.60143
#> 7  19.75238 19.90047 20.03858 20.16738 20.28750 20.39953 20.50400 20.60143
#> 8  19.75238 19.27849 18.83654 18.42438 18.03999 17.68152 17.34720 17.03541
#> 9  19.75238 19.27849 18.83654 18.42438 18.03999 17.68152 17.34720 17.03541
#> 10 19.75238 19.27849 18.83654 18.42438 18.03999 17.68152 17.34720 17.03541
#> 11 19.75238 19.27849 18.83654 18.42438 18.03999 17.68152 17.34720 17.03541
#> 12 19.75238 20.52245 21.24062 21.91038 22.53501 23.11754 23.66081 24.16746
#> 13 19.75238 20.52245 21.24062 21.91038 22.53501 23.11754 23.66081 24.16746
#> 14 19.75238 20.52245 21.24062 21.91038 22.53501 23.11754 23.66081 24.16746
#> 15 19.75238 19.27849 18.83654 18.42438 18.03999 17.68152 17.34720 17.03541
#> 16 19.75238 19.27849 18.83654 18.42438 18.03999 17.68152 17.34720 17.03541
#> 17 19.75238 19.27849 18.83654 18.42438 18.03999 17.68152 17.34720 17.03541
#> 18 19.75238 20.52245 21.24062 21.91038 22.53501 23.11754 23.66081 24.16746
#> 19 19.75238 19.27849 18.83654 18.42438 18.03999 17.68152 17.34720 17.03541
#> 20 19.75238 19.90047 20.03858 20.16738 20.28750 20.39953 20.50400 20.60143
#> 21 19.75238 19.27849 18.83654 18.42438 18.03999 17.68152 17.34720 17.03541
#>      2.8550   2.6626   2.4831   2.3158   2.1597   2.0141   1.8784   1.7518
#> 1  20.69230 20.77704 20.85607 20.92978 20.99852 21.06262 21.12241 21.17816
#> 2  20.69230 20.77704 20.85607 20.92978 20.99852 21.06262 21.12241 21.17816
#> 3  24.63997 25.08063 25.49159 25.87485 26.23229 26.56563 26.87651 27.16644
#> 4  20.69230 20.77704 20.85607 20.92978 20.99852 21.06262 21.12241 21.17816
#> 5  16.74464 16.47346 16.22056 15.98471 15.76475 15.55961 15.36830 15.18989
#> 6  20.69230 20.77704 20.85607 20.92978 20.99852 21.06262 21.12241 21.17816
#> 7  20.69230 20.77704 20.85607 20.92978 20.99852 21.06262 21.12241 21.17816
#> 8  16.74464 16.47346 16.22056 15.98471 15.76475 15.55961 15.36830 15.18989
#> 9  16.74464 16.47346 16.22056 15.98471 15.76475 15.55961 15.36830 15.18989
#> 10 16.74464 16.47346 16.22056 15.98471 15.76475 15.55961 15.36830 15.18989
#> 11 16.74464 16.47346 16.22056 15.98471 15.76475 15.55961 15.36830 15.18989
#> 12 24.63997 25.08063 25.49159 25.87485 26.23229 26.56563 26.87651 27.16644
#> 13 24.63997 25.08063 25.49159 25.87485 26.23229 26.56563 26.87651 27.16644
#> 14 24.63997 25.08063 25.49159 25.87485 26.23229 26.56563 26.87651 27.16644
#> 15 16.74464 16.47346 16.22056 15.98471 15.76475 15.55961 15.36830 15.18989
#> 16 16.74464 16.47346 16.22056 15.98471 15.76475 15.55961 15.36830 15.18989
#> 17 16.74464 16.47346 16.22056 15.98471 15.76475 15.55961 15.36830 15.18989
#> 18 24.63997 25.08063 25.49159 25.87485 26.23229 26.56563 26.87651 27.16644
#> 19 16.74464 16.47346 16.22056 15.98471 15.76475 15.55961 15.36830 15.18989
#> 20 20.69230 20.77704 20.85607 20.92978 20.99852 21.06262 21.12241 21.17816
#> 21 16.74464 16.47346 16.22056 15.98471 15.76475 15.55961 15.36830 15.18989
#>      1.6337   1.5236   1.4209   1.3252   1.2358   1.1526   1.0749   1.0024
#> 1  21.21724 21.21724 21.21724 21.21724 21.21724 21.21724 21.21724 21.21724
#> 2  21.21724 21.21724 21.21724 21.21724 21.21724 21.21724 21.21724 21.21724
#> 3  27.36966 27.36966 27.36966 27.36966 27.36966 27.36966 27.36966 27.36966
#> 4  21.21724 21.21724 21.21724 21.21724 21.21724 21.21724 21.21724 21.21724
#> 5  15.06483 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483
#> 6  21.21724 21.21724 21.21724 21.21724 21.21724 21.21724 21.21724 21.21724
#> 7  21.21724 21.21724 21.21724 21.21724 21.21724 21.21724 21.21724 21.21724
#> 8  15.06483 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483
#> 9  15.06483 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483
#> 10 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483
#> 11 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483
#> 12 27.36966 27.36966 27.36966 27.36966 27.36966 27.36966 27.36966 27.36966
#> 13 27.36966 27.36966 27.36966 27.36966 27.36966 27.36966 27.36966 27.36966
#> 14 27.36966 27.36966 27.36966 27.36966 27.36966 27.36966 27.36966 27.36966
#> 15 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483
#> 16 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483
#> 17 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483
#> 18 27.36966 27.36966 27.36966 27.36966 27.36966 27.36966 27.36966 27.36966
#> 19 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483
#> 20 21.21724 21.21724 21.21724 21.21724 21.21724 21.21724 21.21724 21.21724
#> 21 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483 15.06483
#>      0.9349   0.8719   0.8131   0.7583   0.7072   0.6595   0.6151   0.5736
#> 1  21.21724 22.38338 22.38347 22.38353 22.38356 22.33148 22.26004 22.19342
#> 2  21.21724 21.64392 21.64393 21.64394 21.64394 21.58939 21.51446 21.44457
#> 3  27.36966 26.70710 26.70685 26.70668 26.70657 26.72960 26.76088 26.79003
#> 4  21.21724 20.65797 20.65788 20.65782 20.65778 20.72817 20.82474 20.91481
#> 5  15.06483 16.55175 16.55201 16.55219 16.55230 16.63380 16.74614 16.85093
#> 6  21.21724 19.94751 19.94734 19.94723 19.94716 20.01519 20.10839 20.19532
#> 7  21.21724 20.00551 20.00534 20.00524 20.00517 19.94515 19.86247 19.78534
#> 8  15.06483 15.71079 15.71097 15.71109 15.71116 15.74711 15.79675 15.84305
#> 9  15.06483 15.56580 15.56596 15.56607 15.56614 15.60160 15.65055 15.69622
#> 10 15.06483 11.30301 11.30273 11.30255 11.30243 11.28093 11.25100 11.22306
#> 11 15.06483 11.02753 11.02722 11.02702 11.02688 11.00447 10.97323 10.94408
#> 12 27.36966 27.05508 27.05487 27.05473 27.05463 27.07882 27.11174 27.14243
#> 13 27.36966 28.11353 28.11342 28.11335 28.11331 28.14103 28.17895 28.21432
#> 14 27.36966 26.28662 26.28633 26.28613 26.28600 26.30763 26.33691 26.36421
#> 15 15.06483 16.56624 16.56651 16.56669 16.56680 16.64835 16.76075 16.86561
#> 16 15.06483 15.39180 15.39195 15.39205 15.39211 15.38425 15.37365 15.36378
#> 17 15.06483 15.37731 15.37745 15.37755 15.37761 15.45518 15.56197 15.66157
#> 18 27.36966 29.04728 29.04727 29.04727 29.04726 29.03535 29.01897 29.00370
#> 19 15.06483 17.33471 17.33505 17.33528 17.33543 17.33405 17.33265 17.33136
#> 20 21.21724 21.94841 21.94845 21.94847 21.94849 21.80946 21.61853 21.44045
#> 21 15.06483 16.17476 16.17499 16.17514 16.17524 15.99900 15.75723 15.53173
#>      0.5350   0.4989   0.4653   0.4339   0.4047   0.3774   0.3520   0.3283
#> 1  22.13139 22.07345 22.01940 21.96900 22.30754 22.30748 22.30743 22.30740
#> 2  21.37955 21.31877 21.26207 21.20919 21.87883 21.87873 21.87864 21.87856
#> 3  26.81733 26.84268 26.86633 26.88838 28.39624 28.39601 28.39579 28.39559
#> 4  20.99865 21.07699 21.15006 21.21820 20.90409 20.90413 20.90417 20.90420
#> 5  16.94822 17.03937 17.12440 17.20370 17.18434 17.18422 17.18418 17.18418
#> 6  20.27630 20.35191 20.42243 20.48819 20.49219 20.49220 20.49219 20.49218
#> 7  19.71372 19.64663 19.58405 19.52569 17.16152 17.16205 17.16245 17.16275
#> 8  15.88600 15.92628 15.96386 15.99890 15.57536 15.57537 15.57541 15.57546
#> 9  15.73858 15.77831 15.81536 15.84992 15.49130 15.49130 15.49133 15.49138
#> 10 11.19727 11.17297 11.15029 11.12914 11.89850 11.89844 11.89835 11.89824
#> 11 10.91717 10.89181 10.86815 10.84608 11.73878 11.73871 11.73860 11.73848
#> 12 27.17113 27.19783 27.22272 27.24593 28.59799 28.59777 28.59757 28.59740
#> 13 28.24729 28.27806 28.30675 28.33350 29.21163 29.21147 29.21133 29.21122
#> 14 26.38981 26.41355 26.43568 26.45633 24.38503 24.38550 24.38584 24.38609
#> 15 16.96296 17.05417 17.13925 17.21860 17.19274 17.19263 17.19259 17.19259
#> 16 15.35450 15.34591 15.33791 15.33044 14.26899 14.26916 14.26932 14.26945
#> 17 15.75412 15.84076 15.92157 15.99694 16.50345 16.50327 16.50316 16.50309
#> 18 28.98949 28.97620 28.96381 28.95225 28.63154 28.63161 28.63167 28.63171
#> 19 17.32991 17.32880 17.32777 17.32681 19.16283 19.16239 19.16207 19.16185
#> 20 21.27477 21.11987 20.97540 20.84066 19.81250 19.81276 19.81296 19.81310
#> 21 15.32182 15.12569 14.94276 14.77215 14.00462 14.00481 14.00496 14.00507
#>      0.3061   0.2855   0.2663   0.2483   0.2316   0.2160   0.2014   0.1878
#> 1  22.30737 22.30735 22.30734 22.30733 22.35006 22.35013 22.35018 22.35021
#> 2  21.87850 21.87844 21.87840 21.87836 22.26865 22.26881 22.26892 22.26900
#> 3  28.39542 28.39528 28.39516 28.39507 27.73346 27.73375 27.73396 27.73411
#> 4  20.90422 20.90424 20.90426 20.90427 21.09547 21.09541 21.09536 21.09533
#> 5  17.18421 17.18426 17.18430 17.18435 17.14710 17.14693 17.14681 17.14673
#> 6  20.49217 20.49215 20.49214 20.49213 21.27587 21.27590 21.27592 21.27594
#> 7  17.16296 17.16311 17.16322 17.16329 17.63289 17.63264 17.63246 17.63232
#> 8  15.57552 15.57558 15.57563 15.57568 16.23192 16.23178 16.23168 16.23161
#> 9  15.49143 15.49148 15.49153 15.49157 16.52613 16.52602 16.52594 16.52588
#> 10 11.89814 11.89805 11.89796 11.89789 11.82313 11.82337 11.82354 11.82366
#> 11 11.73836 11.73826 11.73816 11.73808 10.97126 10.97151 10.97168 10.97181
#> 12 28.59724 28.59712 28.59702 28.59693 28.97265 28.97294 28.97314 28.97329
#> 13 29.21112 29.21105 29.21099 29.21094 30.40554 30.40573 30.40587 30.40597
#> 14 24.38627 24.38639 24.38647 24.38652 23.70461 23.70435 23.70416 23.70402
#> 15 17.19262 17.19267 17.19271 17.19276 17.45971 17.45955 17.45944 17.45936
#> 16 14.26957 14.26967 14.26975 14.26981 12.83603 12.83576 12.83557 12.83544
#> 17 16.50306 16.50305 16.50305 16.50305 16.09922 16.09917 16.09914 16.09912
#> 18 28.63174 28.63176 28.63177 28.63179 27.29583 27.29574 27.29568 27.29564
#> 19 19.16170 19.16161 19.16155 19.16152 18.78854 18.78870 18.78881 18.78889
#> 20 19.81320 19.81327 19.81331 19.81334 19.40037 19.40029 19.40023 19.40019
#> 21 14.00516 14.00523 14.00528 14.00532 14.78156 14.78153 14.78150 14.78148
#>      0.1752   0.1634   0.1524   0.1421   0.1325   0.1236   0.1153   0.1075
#> 1  22.35023 22.35025 22.35026 22.35027 22.35074 22.35088 22.35101 22.35106
#> 2  22.26906 22.26910 22.26913 22.26915 22.27749 22.29017 22.30196 22.30327
#> 3  27.73422 27.73430 27.73436 27.73440 27.73000 27.72121 27.71297 27.71243
#> 4  21.09530 21.09529 21.09527 21.09527 21.05619 20.99173 20.93162 20.92566
#> 5  17.14667 17.14663 17.14661 17.14659 17.12287 17.08470 17.04912 17.04543
#> 6  21.27595 21.27596 21.27597 21.27597 21.18734 21.03964 20.90182 20.88831
#> 7  17.63222 17.63215 17.63210 17.63206 17.73608 17.91056 18.07340 18.08930
#> 8  16.23156 16.23152 16.23150 16.23148 16.19785 16.14267 16.09122 16.08601
#> 9  16.52584 16.52581 16.52579 16.52577 16.49363 16.44072 16.39138 16.38642
#> 10 11.82375 11.82381 11.82385 11.82389 11.80817 11.78050 11.75465 11.75234
#> 11 10.97190 10.97196 10.97201 10.97205 11.01291 11.07960 11.14179 11.14816
#> 12 28.97339 28.97347 28.97352 28.97356 29.00627 29.05922 29.10857 29.11374
#> 13 30.40604 30.40609 30.40613 30.40615 30.45232 30.52815 30.59886 30.60606
#> 14 23.70392 23.70385 23.70380 23.70377 23.75168 23.83323 23.90937 23.91669
#> 15 17.45930 17.45927 17.45924 17.45922 17.43529 17.39669 17.36070 17.35698
#> 16 12.83535 12.83528 12.83523 12.83520 12.91850 13.05905 13.19022 13.20287
#> 17 16.09910 16.09909 16.09909 16.09908 16.07555 16.03704 16.00113 15.99750
#> 18 27.29562 27.29560 27.29558 27.29557 27.24736 27.16759 27.09318 27.08572
#> 19 18.78895 18.78900 18.78903 18.78905 18.86417 18.98857 19.10461 19.11611
#> 20 19.40015 19.40013 19.40011 19.40010 19.34628 19.25639 19.17252 19.16415
#> 21 14.78146 14.78144 14.78143 14.78142 14.72931 14.64168 14.55991 14.55178
#>      0.1002   0.0935   0.0872   0.0813   0.0758   0.0707   0.0660   0.0615
#> 1  22.35094 22.06451 22.06443 22.59056 22.59022 22.58993 22.58969 22.50320
#> 2  22.30305 21.92926 21.92915 22.34276 22.34247 22.34224 22.34204 22.18706
#> 3  27.71216 27.13909 27.13889 26.79395 26.79398 26.79399 26.79400 26.99710
#> 4  20.92571 21.35422 21.35436 21.46693 21.46695 21.46695 21.46695 21.53395
#> 5  17.04566 17.53654 17.53669 17.73962 17.73953 17.73945 17.73937 17.64803
#> 6  20.88815 20.78934 20.78934 20.60432 20.60445 20.60456 20.60465 20.56544
#> 7  18.08965 18.01958 18.01955 18.27304 18.27293 18.27286 18.27282 18.18053
#> 8  16.08612 15.58405 15.58392 15.34515 15.34519 15.34524 15.34528 15.06982
#> 9  16.38648 16.06224 16.06216 15.80581 15.80592 15.80601 15.80610 15.46190
#> 10 11.75204 12.25237 12.25252 12.57091 12.57087 12.57082 12.57077 12.79633
#> 11 11.14797 10.68375 10.68358 10.49195 10.49192 10.49191 10.49189 10.73974
#> 12 29.11351 28.94317 28.94308 28.80859 28.80860 28.80860 28.80860 28.73170
#> 13 30.60593 31.30865 31.30883 31.33196 31.33209 31.33220 31.33228 31.08018
#> 14 23.91708 24.11123 24.11128 23.91020 23.91035 23.91048 23.91059 24.24084
#> 15 17.35719 17.16998 17.16993 17.29839 17.29821 17.29806 17.29793 17.05557
#> 16 13.20334 12.85891 12.85878 12.65386 12.65381 12.65377 12.65375 12.82773
#> 17 15.99762 16.37788 16.37798 16.64060 16.64044 16.64029 16.64016 16.68038
#> 18 27.08576 27.08998 27.09000 27.08250 27.08249 27.08249 27.08248 27.49135
#> 19 19.11613 19.58859 19.58868 19.36021 19.36033 19.36042 19.36049 19.19994
#> 20 19.16400 18.61300 18.61289 18.78833 18.78825 18.78820 18.78816 18.95648
#> 21 14.55152 15.32367 15.32397 14.90034 14.90099 14.90153 14.90199 14.85272
#>      0.0574   0.0535   0.0499   0.0465   0.0434   0.0405   0.0377   0.0352
#> 1  22.50361 22.50395 22.56284 22.56283 22.56282 22.56281 22.56281 22.45659
#> 2  22.18748 22.18783 22.29352 22.29358 22.29363 22.29368 22.29373 22.25326
#> 3  26.99691 26.99674 26.92386 26.92382 26.92379 26.92376 26.92374 26.70819
#> 4  21.53384 21.53376 21.30326 21.30299 21.30275 21.30252 21.30232 21.20654
#> 5  17.64813 17.64824 17.73558 17.73547 17.73536 17.73526 17.73517 17.77686
#> 6  20.56533 20.56524 20.36003 20.35984 20.35966 20.35950 20.35936 20.30485
#> 7  18.18081 18.18102 17.82187 17.82182 17.82179 17.82177 17.82175 18.02905
#> 8  15.07002 15.07020 15.14342 15.14351 15.14359 15.14367 15.14374 15.10139
#> 9  15.46209 15.46228 15.59658 15.59671 15.59684 15.59695 15.59705 15.61979
#> 10 12.79616 12.79601 12.86899 12.86900 12.86901 12.86902 12.86902 13.00646
#> 11 10.73959 10.73943 10.77233 10.77248 10.77261 10.77274 10.77285 10.74211
#> 12 28.73176 28.73183 28.72654 28.72654 28.72654 28.72654 28.72654 28.65114
#> 13 31.08024 31.08033 31.22561 31.22568 31.22573 31.22578 31.22582 31.27011
#> 14 24.24047 24.24012 24.59993 24.60054 24.60109 24.60159 24.60205 24.67739
#> 15 17.05592 17.05624 17.05602 17.05590 17.05578 17.05568 17.05558 17.06294
#> 16 12.82768 12.82761 12.82085 12.82098 12.82110 12.82121 12.82131 12.49303
#> 17 16.68043 16.68049 16.71860 16.71844 16.71829 16.71815 16.71802 16.77543
#> 18 27.49099 27.49066 27.33878 27.33863 27.33850 27.33837 27.33826 27.45435
#> 19 19.19991 19.19993 19.07092 19.07059 19.07028 19.07000 19.06975 19.19698
#> 20 18.95646 18.95641 19.02223 19.02241 19.02258 19.02273 19.02287 19.22299
#> 21 14.85216 14.85168 14.83823 14.83824 14.83825 14.83826 14.83826 14.79057
#>      0.0328   0.0306   0.0285   0.0266   0.0248   0.0232   0.0216   0.0201
#> 1  22.45642 22.45626 22.45611 22.45596 22.45582 22.45568 22.45554 22.45541
#> 2  22.25319 22.25311 22.25304 22.25297 22.25290 22.25284 22.25277 22.25271
#> 3  26.70795 26.70772 26.70750 26.70729 26.70708 26.70688 26.70670 26.70651
#> 4  21.20634 21.20616 21.20599 21.20582 21.20566 21.20551 21.20536 21.20523
#> 5  17.77694 17.77703 17.77711 17.77719 17.77727 17.77735 17.77743 17.77750
#> 6  20.30470 20.30456 20.30442 20.30429 20.30416 20.30404 20.30393 20.30382
#> 7  18.02906 18.02906 18.02907 18.02907 18.02907 18.02907 18.02906 18.02906
#> 8  15.10137 15.10135 15.10133 15.10131 15.10129 15.10127 15.10125 15.10123
#> 9  15.61987 15.61995 15.62003 15.62010 15.62017 15.62023 15.62029 15.62035
#> 10 13.00664 13.00682 13.00699 13.00715 13.00731 13.00747 13.00761 13.00776
#> 11 10.74212 10.74213 10.74213 10.74214 10.74214 10.74214 10.74214 10.74215
#> 12 28.65105 28.65097 28.65089 28.65081 28.65073 28.65066 28.65059 28.65053
#> 13 31.27024 31.27037 31.27049 31.27061 31.27072 31.27084 31.27094 31.27105
#> 14 24.67782 24.67822 24.67861 24.67897 24.67932 24.67966 24.67998 24.68029
#> 15 17.06290 17.06286 17.06283 17.06280 17.06277 17.06274 17.06271 17.06269
#> 16 12.49275 12.49248 12.49222 12.49197 12.49172 12.49149 12.49126 12.49105
#> 17 16.77550 16.77556 16.77563 16.77569 16.77576 16.77582 16.77589 16.77595
#> 18 27.45441 27.45447 27.45453 27.45459 27.45465 27.45470 27.45476 27.45481
#> 19 19.19700 19.19702 19.19704 19.19707 19.19709 19.19712 19.19715 19.19718
#> 20 19.22319 19.22337 19.22355 19.22372 19.22387 19.22402 19.22417 19.22430
#> 21 14.79055 14.79053 14.79051 14.79049 14.79048 14.79046 14.79045 14.79044
#>      0.0188   0.0175   0.0163   0.0152   0.0142   0.0133   0.0124   0.0115
#> 1  22.45529 22.34697 22.34691 22.34686 22.34681 22.34676 22.34672 22.34668
#> 2  22.25264 22.19777 22.19777 22.19776 22.19776 22.19775 22.19775 22.19775
#> 3  26.70634 26.80404 26.80420 26.80434 26.80449 26.80463 26.80477 26.80490
#> 4  21.20509 21.10414 21.10407 21.10400 21.10393 21.10387 21.10382 21.10376
#> 5  17.77758 17.74261 17.74248 17.74235 17.74223 17.74211 17.74199 17.74188
#> 6  20.30371 20.31915 20.31924 20.31934 20.31943 20.31952 20.31961 20.31969
#> 7  18.02905 17.96493 17.96509 17.96525 17.96541 17.96556 17.96570 17.96584
#> 8  15.10122 15.18158 15.18172 15.18186 15.18199 15.18211 15.18223 15.18235
#> 9  15.62041 15.70288 15.70299 15.70308 15.70318 15.70326 15.70335 15.70343
#> 10 13.00789 12.94406 12.94387 12.94370 12.94352 12.94336 12.94320 12.94304
#> 11 10.74215 10.83785 10.83796 10.83806 10.83816 10.83826 10.83835 10.83844
#> 12 28.65046 28.69889 28.69896 28.69904 28.69912 28.69919 28.69926 28.69932
#> 13 31.27115 31.22012 31.21995 31.21979 31.21964 31.21949 31.21935 31.21921
#> 14 24.68059 24.68575 24.68556 24.68537 24.68519 24.68501 24.68484 24.68467
#> 15 17.06266 17.09637 17.09646 17.09654 17.09661 17.09669 17.09676 17.09683
#> 16 12.49084 12.45995 12.45996 12.45997 12.45998 12.46000 12.46001 12.46003
#> 17 16.77601 16.76050 16.76041 16.76031 16.76022 16.76013 16.76005 16.75996
#> 18 27.45486 27.49980 27.49984 27.49987 27.49991 27.49994 27.49997 27.49999
#> 19 19.19721 19.30924 19.30934 19.30943 19.30951 19.30959 19.30967 19.30975
#> 20 19.22443 19.25046 19.25052 19.25057 19.25063 19.25068 19.25072 19.25077
#> 21 14.79043 14.67293 14.67271 14.67250 14.67229 14.67209 14.67190 14.67171
#>      0.0107   0.0100   0.0093   0.0087   0.0081   0.0076   0.0071   0.0066
#> 1  22.34664 22.34660 22.34657 22.34653 22.34650 22.34647 22.34645 22.34642
#> 2  22.19775 22.19775 22.19775 22.19775 22.19775 22.19775 22.19775 22.19775
#> 3  26.80503 26.80515 26.80527 26.80539 26.80551 26.80562 26.80572 26.80583
#> 4  21.10371 21.10367 21.10362 21.10358 21.10354 21.10350 21.10346 21.10343
#> 5  17.74177 17.74166 17.74156 17.74146 17.74137 17.74127 17.74118 17.74110
#> 6  20.31977 20.31986 20.31993 20.32001 20.32008 20.32016 20.32023 20.32029
#> 7  17.96598 17.96611 17.96623 17.96636 17.96648 17.96659 17.96670 17.96681
#> 8  15.18246 15.18257 15.18267 15.18277 15.18287 15.18297 15.18306 15.18314
#> 9  15.70351 15.70358 15.70366 15.70373 15.70379 15.70386 15.70392 15.70397
#> 10 12.94289 12.94274 12.94260 12.94246 12.94232 12.94219 12.94207 12.94194
#> 11 10.83853 10.83861 10.83869 10.83876 10.83884 10.83891 10.83898 10.83904
#> 12 28.69939 28.69945 28.69951 28.69957 28.69963 28.69969 28.69974 28.69979
#> 13 31.21907 31.21894 31.21881 31.21869 31.21857 31.21845 31.21834 31.21823
#> 14 24.68451 24.68434 24.68419 24.68403 24.68388 24.68373 24.68359 24.68345
#> 15 17.09690 17.09696 17.09703 17.09709 17.09715 17.09720 17.09726 17.09731
#> 16 12.46005 12.46006 12.46008 12.46010 12.46012 12.46014 12.46016 12.46018
#> 17 16.75988 16.75981 16.75973 16.75966 16.75959 16.75952 16.75945 16.75939
#> 18 27.50002 27.50004 27.50007 27.50009 27.50011 27.50013 27.50015 27.50017
#> 19 19.30982 19.30989 19.30996 19.31002 19.31008 19.31014 19.31020 19.31026
#> 20 19.25081 19.25085 19.25089 19.25092 19.25095 19.25099 19.25101 19.25104
#> 21 14.67153 14.67136 14.67119 14.67103 14.67088 14.67073 14.67058 14.67044
#>      0.0062   0.0057   0.0053   0.0050
#> 1  22.34640 22.34638 22.34636 22.34634
#> 2  22.19775 22.19775 22.19775 22.19776
#> 3  26.80593 26.80603 26.80612 26.80622
#> 4  21.10340 21.10337 21.10334 21.10332
#> 5  17.74101 17.74093 17.74085 17.74078
#> 6  20.32036 20.32043 20.32049 20.32055
#> 7  17.96692 17.96702 17.96712 17.96721
#> 8  15.18323 15.18331 15.18339 15.18346
#> 9  15.70403 15.70408 15.70413 15.70418
#> 10 12.94182 12.94171 12.94160 12.94149
#> 11 10.83910 10.83916 10.83922 10.83928
#> 12 28.69984 28.69989 28.69994 28.69999
#> 13 31.21812 31.21802 31.21792 31.21783
#> 14 24.68331 24.68318 24.68304 24.68292
#> 15 17.09737 17.09742 17.09747 17.09751
#> 16 12.46020 12.46022 12.46024 12.46026
#> 17 16.75933 16.75927 16.75921 16.75915
#> 18 27.50019 27.50020 27.50022 27.50024
#> 19 19.31031 19.31036 19.31041 19.31046
#> 20 19.25107 19.25109 19.25112 19.25114
#> 21 14.67031 14.67018 14.67006 14.66994
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 22.8 21.4 18.7 18.1 17.8 17.3 15.2 10.4 14.7 32.4 33.9 21.5 15.2
#> [16] 13.3 19.2 30.4 15.8 19.7 15.0
#> 
#> $X
#>               am        carb        cyl       disp       drat       gear
#>  [1,]  1.1547005  0.57814716 -0.2936101 -0.6966731  0.8570895  0.4291975
#>  [2,]  1.1547005  0.57814716 -0.2936101 -0.6966731  0.8570895  0.4291975
#>  [3,]  1.1547005 -1.07745608 -1.5267726 -1.1293696  0.7436035  0.4291975
#>  [4,] -0.8660254 -1.07745608 -0.2936101  0.1187934 -1.0040809 -0.8583951
#>  [5,] -0.8660254 -0.52558833  0.9395524  0.9675443 -0.8452005 -0.8583951
#>  [6,] -0.8660254 -1.07745608 -0.2936101 -0.1558024 -1.7303912 -0.8583951
#>  [7,] -0.8660254  0.57814716 -0.2936101 -0.6334328  0.9024839  0.4291975
#>  [8,] -0.8660254  0.02627942  0.9395524  0.2669088 -1.0267781 -0.8583951
#>  [9,] -0.8660254  0.02627942  0.9395524  0.2669088 -1.0267781 -0.8583951
#> [10,] -0.8660254  0.57814716  0.9395524  1.8995060 -1.3445388 -0.8583951
#> [11,] -0.8660254  0.57814716  0.9395524  1.6332312 -0.6636229 -0.8583951
#> [12,]  1.1547005 -1.07745608 -1.5267726 -1.3731775  1.2656390  0.4291975
#> [13,]  1.1547005 -1.07745608 -1.5267726 -1.4364177  1.5833998  0.4291975
#> [14,] -0.8660254 -1.07745608 -1.5267726 -1.0286845  0.4031455 -0.8583951
#> [15,] -0.8660254 -0.52558833  0.9395524  0.5015634 -0.8452005 -0.8583951
#> [16,] -0.8660254  0.57814716  0.9395524  0.8843334  0.4712371 -0.8583951
#> [17,] -0.8660254 -0.52558833  0.9395524  1.3003878 -1.0040809 -0.8583951
#> [18,]  1.1547005 -0.52558833 -1.5267726 -1.2367116  0.5620259  1.7167902
#> [19,]  1.1547005  0.57814716  0.9395524  0.8926545  1.5833998  1.7167902
#> [20,]  1.1547005  1.68188266 -0.2936101 -0.8214894  0.2215679  1.7167902
#> [21,]  1.1547005  2.78561816  0.9395524  0.4766002  0.0399903  1.7167902
#>               hp         qsec         vs          wt
#>  [1,] -0.6962427 -0.674411645 -0.7844645 -0.66936795
#>  [2,] -0.6962427 -0.339059385 -0.7844645 -0.39570286
#>  [3,] -0.9457996  0.613101495  1.2747549 -0.99132688
#>  [4,] -0.6962427  1.110141451  1.2747549 -0.03081607
#>  [5,]  0.2579454 -0.339059385 -0.7844645  0.21065313
#>  [6,] -0.7696418  1.577239241  1.2747549  0.23211706
#>  [7,] -0.5054051  0.786766058  1.2747549  0.21065313
#>  [8,]  0.3313444  0.008269741 -0.7844645  0.52188009
#>  [9,]  0.3313444  0.247807069 -0.7844645  0.57553992
#> [10,]  0.6983399  0.235830203 -0.7844645  2.15313867
#> [11,]  1.0653353 -0.099522057 -0.7844645  2.25509233
#> [12,] -1.3421547  1.128106751  1.2747549 -1.12011045
#> [13,] -1.3568345  1.385609379  1.2747549 -1.51182715
#> [14,] -0.8870804  1.451482144  1.2747549 -0.83571339
#> [15,] -0.1090501 -0.171383255 -0.7844645  0.20528715
#> [16,]  1.2855325 -1.303197131 -0.7844645  0.63993170
#> [17,]  0.2579454 -0.321094085 -0.7844645  0.64529768
#> [18,] -0.6522033 -0.410920583  1.2747549 -1.85739640
#> [19,]  1.5644490 -1.848144553 -0.7844645 -0.07910991
#> [20,]  0.2579454 -1.249301232 -0.7844645 -0.50838848
#> [21,]  2.6067160 -1.788260221 -0.7844645  0.35016866
#> attr(,"center")
#>  [1]   0.4285714   2.9523810   6.4761905 243.7238095   3.5223810   3.6666667
#>  [7] 157.4285714  17.5861905   0.3809524   3.2437143
#> attr(,"scale")
#>  [1]   0.4948717   1.8120283   1.6218463 120.1766078   0.4405830   0.7766432
#>  [7]  68.1207411   1.6698859   0.4856209   0.9317959
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 35
#> 
#> $lambda.min
#> [1] 0.4652884
#> 
#> $null.dev
#> [1] 35.23392
#> 
#> $Bias
#> [1] 11.72991
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 2.605317