Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 35.549114 33.108532 29.146887 24.853731 21.350470 18.518670 16.256372
#>   [8] 14.475873 13.101808 12.069475 11.323394 10.816050 10.506808 10.360965
#>  [15] 10.348942 10.445570 10.494070 10.548512 10.650785 10.742929 10.845592
#>  [22] 10.950639 10.956410 10.932318 10.931297 10.949005 10.899888 10.758434
#>  [29] 10.643679 10.551590 10.403606 10.267725 10.776767 10.898636 10.389432
#>  [36]  8.992082  8.630551  8.304509  9.375783 11.102086 11.038747 10.321476
#>  [43]  9.878196  9.826420  9.825624 12.043935 11.310992 12.171628 12.714811
#>  [50] 14.246348 14.225424 14.591495 14.627803 14.604306 14.546787 14.087627
#>  [57] 14.087824 13.403067 13.271972 12.905692 12.905810 12.968519 12.968565
#>  [64] 12.760197 14.612820 14.556516 14.558064 14.606817 14.464070 14.692590
#>  [71] 14.600700 14.932919 14.934697 14.936548 15.029186 15.030883 15.032537
#>  [78] 14.824519 14.825992 14.831528 14.832922 14.834286 14.835619 14.836923
#>  [85] 14.838197 14.839443 14.840660 14.841850 14.843011 14.844145 14.845252
#>  [92] 14.846333 14.847388 14.848417 14.849444 14.848678 14.849670 14.850639
#>  [99] 14.851585 14.852508
#> 
#> $cvse
#>   [1] 9.274479 9.063828 8.211990 7.058241 6.142425 5.426601 4.878105 4.469006
#>   [9] 4.175651 3.978202 3.860127 3.807658 3.809271 3.855224 3.937210 4.048098
#>  [17] 4.061205 3.992133 3.968776 3.968454 3.968350 3.968401 3.973948 3.984021
#>  [25] 3.995253 4.007303 3.998963 3.975662 3.957686 3.944223 3.876468 3.808767
#>  [33] 3.701635 3.602764 3.238299 2.428284 2.386646 2.348320 2.625404 2.644944
#>  [41] 2.641191 2.639979 2.579665 2.577099 2.570864 3.811701 3.783654 4.209463
#>  [49] 4.297720 4.862722 4.865636 4.922372 4.901684 4.886453 4.878283 4.832849
#>  [57] 4.832935 4.718642 4.827144 4.757341 4.757372 4.763054 4.763074 4.765009
#>  [65] 6.614821 6.621363 6.622686 6.616848 6.596769 6.580237 6.590161 6.612748
#>  [73] 6.613849 6.614910 6.628595 6.629783 6.630947 6.638737 6.639854 6.640547
#>  [81] 6.641618 6.642664 6.643687 6.644687 6.645663 6.646617 6.647548 6.648457
#>  [89] 6.649343 6.650208 6.651052 6.651874 6.652676 6.653457 6.654218 6.654503
#>  [97] 6.655226 6.655930 6.656615 6.657281
#> 
#> $fold
#>  [1] 10  1  5  3  7  5  8  1  1  6  8 10  2  3  4  4  2  7  6  9  9
#> 
#> $lambda
#>   [1] 5.307983364 4.950243051 4.616613237 4.305468956 4.015294758 3.744677330
#>   [7] 3.492298611 3.256929373 3.037423234 2.832711074 2.641795828 2.463747631
#>  [13] 2.297699287 2.142842045 1.998421663 1.863734731 1.738125248 1.620981424
#>  [19] 1.511732701 1.409846977 1.314828009 1.226213002 1.143570349 1.066497535
#>  [25] 0.994619171 0.927585168 0.865069032 0.806766274 0.752392928 0.701684162
#>  [31] 0.654392998 0.610289100 0.569157658 0.530798336 0.495024305 0.461661324
#>  [37] 0.430546896 0.401529476 0.374467733 0.349229861 0.325692937 0.303742323
#>  [43] 0.283271107 0.264179583 0.246374763 0.229769929 0.214284205 0.199842166
#>  [49] 0.186373473 0.173812525 0.162098142 0.151173270 0.140984698 0.131482801
#>  [55] 0.122621300 0.114357035 0.106649754 0.099461917 0.092758517 0.086506903
#>  [61] 0.080676627 0.075239293 0.070168416 0.065439300 0.061028910 0.056915766
#>  [67] 0.053079834 0.049502431 0.046166132 0.043054690 0.040152948 0.037446773
#>  [73] 0.034922986 0.032569294 0.030374232 0.028327111 0.026417958 0.024637476
#>  [79] 0.022976993 0.021428420 0.019984217 0.018637347 0.017381252 0.016209814
#>  [85] 0.015117327 0.014098470 0.013148280 0.012262130 0.011435703 0.010664975
#>  [91] 0.009946192 0.009275852 0.008650690 0.008067663 0.007523929 0.007016842
#>  [97] 0.006543930 0.006102891 0.005691577 0.005307983
#> 
#> $fit
#> $beta
#>               5.3080     4.9502    4.6166    4.3055    4.0153   3.7447
#> (Intercept) 20.52857 22.3704376 24.088168 25.690129 27.184124 28.57743
#> am           0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> carb         0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> cyl          0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> disp         0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> drat         0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> gear         0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> hp           0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> qsec         0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> vs           0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> wt           0.00000 -0.5987213 -1.157091 -1.677828 -2.163469 -2.61638
#>                3.4923    3.2569    3.0374    2.8327    2.6418    2.4637
#> (Intercept) 29.876828 31.088653 32.218805 33.272789 34.255737 35.172438
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -3.038766 -3.432684 -3.800054 -4.142665 -4.462184 -4.760169
#>                2.2977    2.1428    1.9984    1.8637    1.7381    1.6210
#> (Intercept) 36.027357 36.824657 37.568221 38.261672 38.747737 38.747737
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -5.038071 -5.297243 -5.538948 -5.764363 -5.922364 -5.922364
#>                1.5117    1.4098    1.3148      1.2262     1.1436     1.0665
#> (Intercept) 38.747737 38.747737 38.747737 38.10924242 36.5184485 35.0347866
#> am           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> carb         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> cyl          0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> disp         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> qsec         0.000000  0.000000  0.000000  0.03446475  0.1203328  0.2004182
#> vs           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> wt          -5.922364 -5.922364 -5.922364 -5.91247503 -5.8878371 -5.8648585
#>                0.9946     0.9276     0.8651     0.8068    0.7524     0.7017
#> (Intercept) 33.651119 32.3607051 31.1572612 30.0349254 28.988231 28.0120808
#> am           0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> carb         0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> cyl          0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> disp         0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> drat         0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> gear         0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> hp           0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> qsec         0.275106  0.3447601  0.4097198  0.4703014  0.526800  0.5794908
#> vs           0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> wt          -5.843428 -5.8234428 -5.8048041 -5.7874216 -5.771211 -5.7560922
#>                 0.6544     0.6103     0.5692     0.5308     0.4950    0.4617
#> (Intercept) 27.1017196 26.2527138 25.4609280 24.7225060 24.0338511 11.195591
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  1.202783
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> qsec         0.6286304  0.6744581  0.7171973  0.7570559  0.7942282  1.140171
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> wt          -5.7419927 -5.7288435 -5.7165805 -5.7051439 -5.6944782 -4.957497
#>                0.4305    0.4015    0.3745    0.3492     0.3257     0.3037
#> (Intercept)  6.387845  6.384148  6.381835  6.380388  4.6669289  3.2625753
#> am           0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> carb         0.000000  0.000000  0.000000  0.000000 -1.0575133 -1.0866235
#> cyl          0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.000000  1.3709514  1.7298791
#> gear         1.699877  1.700233  1.700456  1.700595  2.8536558  2.8340409
#> hp           0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> qsec         1.256256  1.256353  1.256413  1.256451  0.6863092  0.6703851
#> vs           0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> wt          -4.660627 -4.660411 -4.660275 -4.660190 -2.8591613 -2.6798114
#>                0.2833     0.2642     0.2464     0.2298     0.2143     0.1998
#> (Intercept)  3.261231  3.2600962  3.2591387  3.2583318  3.2576522  3.2570802
#> am           0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb        -1.086874 -1.0870837 -1.0872602 -1.0874086 -1.0875333 -1.0876382
#> cyl          0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         1.730057  1.7302070  1.7303334  1.7304397  1.7305291  1.7306043
#> gear         2.834481  2.8348505  2.8351614  2.8354229  2.8356428  2.8358277
#> hp           0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.670293  0.6702156  0.6701507  0.6700962  0.6700504  0.6700119
#> vs           0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.679356 -2.6789733 -2.6786513 -2.6783805 -2.6781528 -2.6779613
#>                 0.1864     0.1738     0.1621     0.1512     0.1410     0.1315
#> (Intercept)  3.2565990  3.2561943  3.2558539 -40.028395 -40.034887 -40.041007
#> am           0.0000000  0.0000000  0.0000000   0.000000   0.000000   0.000000
#> carb        -1.0877264 -1.0878005 -1.0878628  -1.449531  -1.449607  -1.449680
#> cyl          0.0000000  0.0000000  0.0000000   1.890202   1.890500   1.890781
#> disp         0.0000000  0.0000000  0.0000000   0.000000   0.000000   0.000000
#> drat         1.7306675  1.7307206  1.7307653   3.529400   3.529612   3.529812
#> gear         2.8359832  2.8361140  2.8362240   5.033600   5.033989   5.034356
#> hp           0.0000000  0.0000000  0.0000000   0.000000   0.000000   0.000000
#> qsec         0.6699795  0.6699523  0.6699294   1.797652   1.797820   1.797978
#> vs           0.0000000  0.0000000  0.0000000   0.000000   0.000000   0.000000
#> wt          -2.6778003 -2.6776649 -2.6775511  -3.241443  -3.241534  -3.241620
#>                 0.1226     0.1144      0.1066       0.0995       0.0928
#> (Intercept) -40.046775 -40.052213 -33.1505291 -33.15261415 -33.15462672
#> am            0.000000   0.000000   0.0000000   0.00000000   0.00000000
#> carb         -1.449748  -1.449812  -1.3202871  -1.32030408  -1.32032127
#> cyl           1.891046   1.891295   2.4186312   2.41892550   2.41920146
#> disp          0.000000   0.000000   0.0000000   0.00000000   0.00000000
#> drat          3.530000   3.530178   3.7161300   3.71625554   3.71637485
#> gear          5.034702   5.035028   5.1559399   5.15616090   5.15637053
#> hp            0.000000   0.000000  -0.0184378  -0.01844112  -0.01844418
#> qsec          1.798128   1.798268   1.0906225   1.09057988   1.09054238
#> vs            0.000000   0.000000   2.7621166   2.76266407   2.76317144
#> wt           -3.241702  -3.241778  -2.4512195  -2.45119370  -2.45117064
#>                 0.0865       0.0807       0.0752       0.0702      0.0654
#> (Intercept) -33.156567 -33.15843626 -33.16023486 -33.16196411 -33.1636253
#> am            0.000000   0.00000000   0.00000000   0.00000000   0.0000000
#> carb         -1.320339  -1.32035579  -1.32037295  -1.32038995  -1.3204067
#> cyl           2.419460   2.41970296   2.41993060   2.42014412   2.4203444
#> disp          0.000000   0.00000000   0.00000000   0.00000000   0.0000000
#> drat          3.716488   3.71659581   3.71669801   3.71679502   3.7168871
#> gear          5.156569   5.15675793   5.15693671   5.15710619   5.1572668
#> hp           -0.018447  -0.01844959  -0.01845198  -0.01845418  -0.0184562
#> qsec          1.090510   1.09048086   1.09045597   1.09043446   1.0904160
#> vs            2.763642   2.76407752   2.76448154   2.76485607   2.7652033
#> wt           -2.451150  -2.45113177  -2.45111550  -2.45110109  -2.4510883
#>                   0.0610       0.0569       0.0531       0.0495       0.0462
#> (Intercept) -33.16521986 -33.16674934 -33.16821537 -33.16961965 -33.17096395
#> am            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> carb         -1.32042326  -1.32043949  -1.32045538  -1.32047090  -1.32048604
#> cyl           2.42053233   2.42070862   2.42087402   2.42102923   2.42117486
#> disp          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> drat          3.71697443   3.71705730   3.71713590   3.71721042   3.71728108
#> gear          5.15741903   5.15756325   5.15769988   5.15782928   5.15795183
#> hp           -0.01845807  -0.01845979  -0.01846137  -0.01846282  -0.01846416
#> qsec          1.09040023   1.09038691   1.09037576   1.09036656   1.09035908
#> vs            2.76552526   2.76582380   2.76610066   2.76635742   2.76659558
#> wt           -2.45107710  -2.45106723  -2.45105858  -2.45105105  -2.45104450
#>                   0.0431       0.0402       0.0374       0.0349       0.0326
#> (Intercept) -33.17225007 -33.17347984 -33.17465510 -33.17577771 -33.17684950
#> am            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> carb         -1.32050078  -1.32051509  -1.32052898  -1.32054244  -1.32055546
#> cyl           2.42131154   2.42143981   2.42156020   2.42167319   2.42177926
#> disp          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> drat          3.71734806   3.71741154   3.71747170   3.71752869   3.71758267
#> gear          5.15806788   5.15817775   5.15828175   5.15838020   5.15847336
#> hp           -0.01846539  -0.01846653  -0.01846758  -0.01846854  -0.01846942
#> qsec          1.09035314   1.09034855   1.09034517   1.09034284   1.09034144
#> vs            2.76681650   2.76702144   2.76721158   2.76738801   2.76755172
#> wt           -2.45103885  -2.45103401  -2.45102988  -2.45102640  -2.45102349
#>                   0.0304       0.0283       0.0264       0.0246      0.0230
#> (Intercept) -33.17787230 -33.17884793 -33.17977818 -33.18066479 -33.1815095
#> am            0.00000000   0.00000000   0.00000000   0.00000000   0.0000000
#> carb         -1.32056803  -1.32058017  -1.32059188  -1.32060315  -1.3206140
#> cyl           2.42187884   2.42197231   2.42206007   2.42214247   2.4222198
#> disp          0.00000000   0.00000000   0.00000000   0.00000000   0.0000000
#> drat          3.71763381   3.71768223   3.71772808   3.71777149   3.7178126
#> gear          5.15856153   5.15864494   5.15872386   5.15879851   5.1588691
#> hp           -0.01847024  -0.01847099  -0.01847168  -0.01847231  -0.0184729
#> qsec          1.09034086   1.09034099   1.09034173   1.09034301   1.0903447
#> vs            2.76770365   2.76784467   2.76797556   2.76809706   2.7682099
#> wt           -2.45102109  -2.45101915  -2.45101762  -2.45101644  -2.4510156
#>                   0.0214       0.0200        0.0186        0.0174        0.0162
#> (Intercept) -33.18231399 -33.18307991 -33.829738117 -33.826543780 -33.823458979
#> am            0.00000000   0.00000000   0.000000000   0.000000000   0.000000000
#> carb         -1.32062442  -1.32063444  -1.229010953  -1.228963480  -1.228919231
#> cyl           2.42229249   2.42236071   2.347829322   2.347519659   2.347223316
#> disp          0.00000000   0.00000000   0.003318709   0.003318748   0.003318735
#> drat          3.71785147   3.71788828   3.653995030   3.653782228   3.653579651
#> gear          5.15893591   5.15899906   5.123313711   5.123012361   5.122723541
#> hp           -0.01847344  -0.01847393  -0.019039314  -0.019036096  -0.019033016
#> qsec          1.09034686   1.09034931   1.189375615   1.189404025   1.189429050
#> vs            2.76831460   2.76841186   2.617476957   2.616945902   2.616439816
#> wt           -2.45101500  -2.45101467  -2.813679131  -2.813782338  -2.813874244
#>                    0.0151        0.0141        0.0131        0.0123
#> (Intercept) -33.820481264 -33.817608179 -33.814837264 -33.812166064
#> am            0.000000000   0.000000000   0.000000000   0.000000000
#> carb         -1.228878055  -1.228839810  -1.228804357  -1.228771561
#> cyl           2.346939870   2.346668904   2.346410010   2.346162786
#> disp          0.003318675   0.003318571   0.003318425   0.003318241
#> drat          3.653386932   3.653203710   3.653029638   3.652864372
#> gear          5.122446867   5.122181962   5.121928455   5.121685982
#> hp           -0.019030069  -0.019027251  -0.019024558  -0.019021986
#> qsec          1.189450879   1.189469694   1.189485670   1.189498975
#> vs            2.615957827   2.615499080   2.615062743   2.614648002
#> wt           -2.813955472  -2.814026621  -2.814088267  -2.814140961
#>                    0.0114        0.0107       0.0099        0.0093
#> (Intercept) -33.809592128 -33.807113013 -33.80472629 -33.802429544
#> am            0.000000000   0.000000000   0.00000000   0.000000000
#> carb         -1.228741293  -1.228713429  -1.22868785  -1.228664439
#> cyl           2.345926841   2.345701791   2.34548726   2.345282880
#> disp          0.003318023   0.003317771   0.00331749   0.003317182
#> drat          3.652707582   3.652558944   3.65241814   3.652284873
#> gear          5.121454184   5.121232713   5.12102122   5.120819379
#> hp           -0.019019531  -0.019017189  -0.01901496  -0.019012829
#> qsec          1.189509769   1.189518205   1.18952443   1.189528586
#> vs            2.614254064   2.613880154   2.61352552   2.613189425
#> wt           -2.814185230  -2.814221580  -2.81425050  -2.814272441
#>                    0.0087        0.0081        0.0075        0.0070
#> (Intercept) -33.800220381 -33.798096425 -33.796055323 -33.794094751
#> am            0.000000000   0.000000000   0.000000000   0.000000000
#> carb         -1.228643086  -1.228623685  -1.228606133  -1.228590332
#> cyl           2.345088293   2.344903148   2.344727104   2.344559826
#> disp          0.003316849   0.003316494   0.003316118   0.003315723
#> drat          3.652158837   3.652039746   3.651927320   3.651821285
#> gear          5.120626851   5.120443318   5.120268465   5.120101984
#> hp           -0.019010803  -0.019008875  -0.019007041  -0.019005299
#> qsec          1.189530808   1.189531222   1.189529954   1.189527119
#> vs            2.612871155   2.612570015   2.612285328   2.612016437
#> wt           -2.814287857  -2.814297169  -2.814300780  -2.814299078
#>                    0.0065        0.0061        0.0057        0.0053
#> (Intercept) -33.792212408 -33.790406026 -33.788673366 -33.787012221
#> am            0.000000000   0.000000000   0.000000000   0.000000000
#> carb         -1.228576186  -1.228563605  -1.228552502  -1.228542793
#> cyl           2.344400989   2.344250277   2.344107379   2.343971998
#> disp          0.003315313   0.003314887   0.003314449   0.003313999
#> drat          3.651721379   3.651627345   3.651538934   3.651455907
#> gear          5.119943575   5.119792946   5.119649810   5.119513890
#> hp           -0.019003645  -0.019002075  -0.019000586  -0.018999175
#> qsec          1.189522830   1.189517193   1.189510309   1.189502277
#> vs            2.611762704   2.611523509   2.611298250   2.611086344
#> wt           -2.814292429  -2.814281186  -2.814265684  -2.814246242
#> 
#> $iter
#>   [1]   0   4   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   4   3   3   3   3   3   3   3   3   3   3   3   3   3  52
#>  [37]  25   1   1   1 218  24   1   1   1   1   1   1   1   1   1 107   1   1
#>  [55]   1   1 124   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1 118   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.307983364 4.950243051 4.616613237 4.305468956 4.015294758 3.744677330
#>   [7] 3.492298611 3.256929373 3.037423234 2.832711074 2.641795828 2.463747631
#>  [13] 2.297699287 2.142842045 1.998421663 1.863734731 1.738125248 1.620981424
#>  [19] 1.511732701 1.409846977 1.314828009 1.226213002 1.143570349 1.066497535
#>  [25] 0.994619171 0.927585168 0.865069032 0.806766274 0.752392928 0.701684162
#>  [31] 0.654392998 0.610289100 0.569157658 0.530798336 0.495024305 0.461661324
#>  [37] 0.430546896 0.401529476 0.374467733 0.349229861 0.325692937 0.303742323
#>  [43] 0.283271107 0.264179583 0.246374763 0.229769929 0.214284205 0.199842166
#>  [49] 0.186373473 0.173812525 0.162098142 0.151173270 0.140984698 0.131482801
#>  [55] 0.122621300 0.114357035 0.106649754 0.099461917 0.092758517 0.086506903
#>  [61] 0.080676627 0.075239293 0.070168416 0.065439300 0.061028910 0.056915766
#>  [67] 0.053079834 0.049502431 0.046166132 0.043054690 0.040152948 0.037446773
#>  [73] 0.034922986 0.032569294 0.030374232 0.028327111 0.026417958 0.024637476
#>  [79] 0.022976993 0.021428420 0.019984217 0.018637347 0.017381252 0.016209814
#>  [85] 0.015117327 0.014098470 0.013148280 0.012262130 0.011435703 0.010664975
#>  [91] 0.009946192 0.009275852 0.008650690 0.008067663 0.007523929 0.007016842
#>  [97] 0.006543930 0.006102891 0.005691577 0.005307983
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 35
#> 
#> $loss
#>   [1] 732.04286 618.46041 523.43180 444.28699 378.72079 324.74420 280.64212
#>   [8] 244.93671 216.35554 193.80397 176.34114 163.15927 153.56551 146.96641
#>  [15] 142.85429 140.79555 140.37442 140.37442 140.37442 140.37442 140.37442
#>  [22] 137.67815 131.44041 126.24008 121.92718 118.37195 115.46251 113.10243
#>  [29] 111.20867 109.70978 108.54434 107.65960 107.01031 106.55770 106.26861
#>  [36]  92.81894  91.57600  91.57600  91.57600  91.57600  73.52501  73.22868
#>  [43]  73.22867  73.22866  73.22866  73.22865  73.22865  73.22865  73.22865
#>  [50]  73.22865  73.22864  62.06294  62.06293  62.06293  62.06292  62.06291
#>  [57]  57.75856  57.75855  57.75855  57.75854  57.75853  57.75853  57.75852
#>  [64]  57.75852  57.75852  57.75851  57.75851  57.75851  57.75851  57.75850
#>  [71]  57.75850  57.75850  57.75850  57.75850  57.75850  57.75850  57.75850
#>  [78]  57.75849  57.75849  57.75849  57.75849  57.60753  57.60752  57.60751
#>  [85]  57.60751  57.60750  57.60749  57.60748  57.60748  57.60747  57.60747
#>  [92]  57.60746  57.60746  57.60745  57.60745  57.60745  57.60745  57.60744
#>  [99]  57.60744  57.60744
#> 
#> $linear.predictors
#>      5.3080   4.9502   4.6166   4.3055   4.0153   3.7447   3.4923   3.2569
#> 1  20.52857 20.80179 21.05659 21.29422 21.51583 21.72251 21.91526 22.09502
#> 2  20.52857 20.64911 20.76153 20.86637 20.96415 21.05534 21.14038 21.21969
#> 3  20.52857 20.29886 20.08463 19.88484 19.69852 19.52475 19.36270 19.21156
#> 4  20.52857 20.46052 20.39705 20.33786 20.28266 20.23118 20.18317 20.13839
#> 5  20.52857 19.93364 19.37881 18.86137 18.37880 17.92876 17.50905 17.11763
#> 6  20.52857 20.13721 19.77222 19.43183 19.11438 18.81833 18.54223 18.28474
#> 7  20.52857 20.10727 19.71437 19.34794 19.00621 18.68751 18.39029 18.11311
#> 8  20.52857 19.22715 18.01344 16.88153 15.82591 14.84143 13.92331 13.06706
#> 9  20.52857 21.05325 21.54257 21.99891 22.42449 22.82139 23.19154 23.53675
#> 10 20.52857 21.40350 22.21947 22.98044 23.69012 24.35197 24.96922 25.54487
#> 11 20.52857 20.89459 21.23594 21.55428 21.85117 22.12805 22.38627 22.62709
#> 12 20.52857 20.26294 20.01521 19.78418 19.56871 19.36777 19.18037 19.00560
#> 13 20.52857 20.31383 20.11356 19.92679 19.75261 19.59016 19.43867 19.29738
#> 14 20.52857 20.07135 19.64494 19.24727 18.87640 18.53053 18.20797 17.90714
#> 15 20.52857 20.06835 19.63915 19.23888 18.86558 18.51745 18.19277 17.88998
#> 16 20.52857 21.21191 21.84920 22.44353 22.99781 23.51473 23.99682 24.44641
#> 17 20.52857 21.08917 21.61199 22.09958 22.55430 22.97838 23.37387 23.74271
#> 18 20.52857 21.46457 22.33749 23.15158 23.91079 24.61885 25.27918 25.89500
#> 19 20.52857 20.71198 20.88303 21.04255 21.19131 21.33006 21.45945 21.58012
#> 20 20.52857 20.23300 19.95735 19.70028 19.46054 19.23695 19.02843 18.83397
#> 21 20.52857 20.70599 20.87146 21.02577 21.16968 21.30389 21.42906 21.54579
#>      3.0374   2.8327   2.6418   2.4637    2.2977    2.1428    1.9984    1.8637
#> 1  22.26266 22.41901 22.56481 22.70080 22.827611 22.945880 23.056178 23.159042
#> 2  21.29365 21.36263 21.42696 21.48695 21.542903 21.595083 21.643746 21.689130
#> 3  19.07062 18.93917 18.81658 18.70225 18.595632 18.496196 18.403462 18.316978
#> 4  20.09663 20.05769 20.02137 19.98750 19.955911 19.926451 19.898978 19.873356
#> 5  16.75258 16.41214 16.09465 15.79855 15.522408 15.264878 15.024704 14.800716
#> 6  18.04460 17.82065 17.61179 17.41701 17.235352 17.065940 16.907946 16.760600
#> 7  17.85460 17.61352 17.38868 17.17900 16.983449 16.801078 16.630998 16.472382
#> 8  12.26852 11.52380 10.82927 10.18155  9.577485  9.014131  8.488745  7.998769
#> 9  23.85869 24.15893 24.43893 24.70007 24.943601 25.170722 25.382536 25.580074
#> 10 26.08172 26.58239 27.04931 27.48477 27.890872 28.269609 28.622821 28.952227
#> 11 22.85167 23.06112 23.25645 23.43862 23.608512 23.766953 23.914715 24.052518
#> 12 18.84261 18.69061 18.54885 18.41664 18.293347 18.178361 18.071125 17.971116
#> 13 19.16562 19.04274 18.92813 18.82126 18.721583 18.628627 18.541935 18.461087
#> 14 17.62660 17.36496 17.12095 16.89339 16.681165 16.483243 16.298662 16.126520
#> 15 17.60760 17.34424 17.09864 16.86959 16.655974 16.456757 16.270967 16.097698
#> 16 24.86570 25.25673 25.62141 25.96151 26.278690 26.574492 26.850357 27.107631
#> 17 24.08669 24.40749 24.70666 24.98568 25.245885 25.488557 25.714873 25.925936
#> 18 26.46932 27.00494 27.50445 27.97030 28.404756 28.809928 29.187793 29.540192
#> 19 21.69265 21.79761 21.89549 21.98677 22.071900 22.151294 22.225336 22.294388
#> 20 18.65261 18.48348 18.32574 18.17863 18.041444 17.913499 17.794177 17.682898
#> 21 21.65465 21.75618 21.85087 21.93917 22.021520 22.098321 22.169946 22.236744
#>       1.7381    1.6210    1.5117    1.4098    1.3148    1.2262    1.1436
#> 1  23.231143 23.231143 23.231143 23.231143 23.231143 23.185848 23.072994
#> 2  21.720941 21.720941 21.720941 21.720941 21.720941 21.697467 21.638982
#> 3  18.256358 18.256358 18.256358 18.256358 18.256358 18.348956 18.579662
#> 4  19.855396 19.855396 19.855396 19.855396 19.855396 19.937742 20.142905
#> 5  14.643716 14.643716 14.643716 14.643716 14.643716 14.645156 14.648743
#> 6  16.657320 16.657320 16.657320 16.657320 16.657320 16.662290 16.674674
#> 7  16.361201 16.361201 16.361201 16.361201 16.361201 16.380452 16.428415
#> 8   7.655326  7.655326  7.655326  7.655326  7.655326  7.688425  7.770888
#> 9  25.718536 25.718536 25.718536 25.718536 25.718536 25.772826 25.908087
#> 10 29.183119 29.183119 29.183119 29.183119 29.183119 29.198882 29.238156
#> 11 24.149110 24.149110 24.149110 24.149110 24.149110 24.224631 24.412790
#> 12 17.901016 17.901016 17.901016 17.901016 17.901016 17.878751 17.823277
#> 13 18.404417 18.404417 18.404417 18.404417 18.404417 18.396131 18.375486
#> 14 16.005860 16.005860 16.005860 16.005860 16.005860 15.936440 15.763483
#> 15 15.976248 15.976248 15.976248 15.976248 15.976248 15.963400 15.931390
#> 16 27.287963 27.287963 27.287963 27.287963 27.287963 27.319987 27.399774
#> 17 26.073878 26.073878 26.073878 26.073878 26.073878 26.032107 25.928036
#> 18 29.787200 29.787200 29.787200 29.787200 29.787200 29.746122 29.643776
#> 19 22.342789 22.342789 22.342789 22.342789 22.342789 22.265890 22.074299
#> 20 17.604898 17.604898 17.604898 17.604898 17.604898 17.504892 17.255730
#> 21 22.283565 22.283565 22.283565 22.283565 22.283565 22.313606 22.388452
#>       1.0665    0.9946    0.9276    0.8651    0.8068    0.7524    0.7017
#> 1  22.967740 22.869581 22.778036 22.692662 22.613042 22.538787 22.469537
#> 2  21.584436 21.533566 21.486124 21.441880 21.400618 21.362137 21.326249
#> 3  18.794832 18.995499 19.182642 19.357173 19.519941 19.671738 19.813305
#> 4  20.334251 20.512702 20.679125 20.834332 20.979078 21.114069 21.239962
#> 5  14.652089 14.655209 14.658119 14.660833 14.663363 14.665724 14.667925
#> 6  16.686224 16.696996 16.707041 16.716410 16.725147 16.733295 16.740895
#> 7  16.473148 16.514867 16.553773 16.590058 16.623897 16.655455 16.684886
#> 8   7.847798  7.919525  7.986417  8.048801  8.106981  8.161239  8.211841
#> 9  26.034240 26.151890 26.261610 26.363936 26.459366 26.548364 26.631363
#> 10 29.274785 29.308945 29.340802 29.370513 29.398221 29.424062 29.448161
#> 11 24.588278 24.751938 24.904568 25.046912 25.179662 25.303465 25.418924
#> 12 17.771539 17.723288 17.678289 17.636323 17.597186 17.560686 17.526646
#> 13 18.356232 18.338275 18.321529 18.305911 18.291346 18.277763 18.265095
#> 14 15.602174 15.451736 15.311438 15.180595 15.058571 14.944770 14.838640
#> 15 15.901535 15.873693 15.847727 15.823512 15.800928 15.779866 15.760224
#> 16 27.474189 27.543588 27.608309 27.668669 27.724961 27.777458 27.826418
#> 17 25.830973 25.740452 25.656031 25.577301 25.503876 25.435400 25.371539
#> 18 29.548323 29.459303 29.376282 29.298857 29.226650 29.159309 29.096507
#> 19 21.895610 21.728965 21.573550 21.428610 21.293439 21.167378 21.049812
#> 20 17.023347 16.806626 16.604512 16.416019 16.240230 16.076289 15.923397
#> 21 22.458258 22.523359 22.584072 22.640694 22.693499 22.742745 22.788673
#>       0.6544    0.6103    0.5692    0.5308    0.4950    0.4617    0.4305
#> 1  22.404955 22.344725 22.288554 22.236169 22.187315 21.785292 21.654481
#> 2  21.292780 21.261566 21.232457 21.205309 21.179991 21.159626 21.169524
#> 3  19.945331 20.068459 20.183288 20.290379 20.390251 20.705255 20.763199
#> 4  21.357371 21.466866 21.568982 21.664215 21.753030 22.995724 23.445069
#> 5  14.669978 14.671892 14.673678 14.675343 14.676896 14.465900 14.377575
#> 6  16.747982 16.754591 16.760755 16.766503 16.771864 16.379483 16.213439
#> 7  16.712334 16.737932 16.761805 16.784068 16.804832 16.587677 16.482910
#> 8   8.259032  8.303043  8.344087  8.382366  8.418064  9.277353  9.606663
#> 9  26.708769 26.780958 26.848282 26.911068 26.969623 27.299355 27.393274
#> 10 29.470636 29.491596 29.511144 29.529374 29.546376 29.116329 28.926298
#> 11 25.526602 25.627022 25.720674 25.808015 25.889469 25.398528 25.136710
#> 12 17.494900 17.465294 17.437683 17.411933 17.387918 16.588233 16.275104
#> 13 18.253280 18.242262 18.231987 18.222404 18.213467 17.499893 17.211448
#> 14 14.739662 14.647355 14.561269 14.480985 14.406112 13.337184 12.949570
#> 15 15.741906 15.724822 15.708890 15.694031 15.680174 15.182277 14.986526
#> 16 27.872078 27.914661 27.954373 27.991409 28.025949 27.963195 27.912275
#> 17 25.311983 25.256440 25.204640 25.156332 25.111279 25.641315 25.892960
#> 18 29.037938 28.983316 28.932376 28.884868 28.840563 28.977699 29.066424
#> 19 20.940171 20.837919 20.742558 20.653624 20.570684 21.149887 21.449257
#> 20 15.780809 15.647831 15.523816 15.408159 15.300296 16.157736 16.590125
#> 21 22.831505 22.871450 22.908703 22.943446 22.975847 23.432059 23.597168
#>       0.4015    0.3745    0.3492   0.3257   0.3037   0.2833   0.2642   0.2464
#> 1  21.654369 21.654299 21.654256 21.00386 21.01221 21.01199 21.01181 21.01166
#> 2  21.169522 21.169520 21.169520 20.65910 20.70427 20.70412 20.70399 20.70389
#> 3  20.763277 20.763326 20.763356 19.93868 19.73558 19.73551 19.73545 19.73540
#> 4  23.445423 23.445645 23.445784 23.63080 23.66785 23.66803 23.66818 23.66831
#> 5  14.377512 14.377473 14.377448 14.56917 14.57342 14.57345 14.57346 14.57348
#> 6  16.213322 16.213249 16.213203 15.67855 15.61864 15.61848 15.61836 15.61825
#> 7  16.482843 16.482801 16.482774 15.81011 15.75280 15.75263 15.75249 15.75238
#> 8   9.606912  9.607068  9.607166 10.34397 10.47126 10.47149 10.47169 10.47185
#> 9  27.393363 27.393418 27.393453 27.68981 27.72684 27.72693 27.72702 27.72709
#> 10 28.926168 28.926087 28.926036 28.81822 29.04143 29.04125 29.04110 29.04098
#> 11 25.136551 25.136453 25.136391 23.92812 23.88730 23.88696 23.88668 23.88644
#> 12 16.274871 16.274725 16.274634 16.41048 16.24238 16.24239 16.24240 16.24241
#> 13 17.211238 17.211106 17.211024 17.48330 17.43308 17.43309 17.43309 17.43309
#> 14 12.949265 12.949074 12.948955 13.70834 13.91081 13.91078 13.91075 13.91073
#> 15 14.986381 14.986290 14.986233 16.04350 16.04567 16.04587 16.04604 16.04618
#> 16 27.912251 27.912236 27.912226 28.05629 28.05487 28.05490 28.05492 28.05494
#> 17 25.893124 25.893226 25.893291 28.23625 28.38353 28.38411 28.38460 28.38501
#> 18 29.066472 29.066501 29.066520 29.26138 29.05613 29.05629 29.05642 29.05653
#> 19 21.449442 21.449557 21.449630 20.27089 20.14309 20.14292 20.14278 20.14266
#> 20 16.590396 16.590566 16.590672 15.14118 15.08426 15.08402 15.08383 15.08366
#> 21 23.597298 23.597379 23.597430 24.41802 24.55458 24.55478 24.55495 24.55509
#>      0.2298   0.2143   0.1998   0.1864   0.1738   0.1621   0.1512   0.1410
#> 1  21.01153 21.01142 21.01133 21.01126 21.01119 21.01114 20.51052 20.51042
#> 2  20.70380 20.70372 20.70366 20.70361 20.70356 20.70352 20.69064 20.69061
#> 3  19.73535 19.73532 19.73529 19.73526 19.73524 19.73522 19.83836 19.83841
#> 4  23.66842 23.66851 23.66859 23.66865 23.66870 23.66875 23.40407 23.40403
#> 5  14.57349 14.57350 14.57351 14.57352 14.57352 14.57353 14.76716 14.76719
#> 6  15.61816 15.61808 15.61802 15.61796 15.61792 15.61788 16.22878 16.22888
#> 7  15.75228 15.75219 15.75212 15.75206 15.75202 15.75197 16.78577 16.78593
#> 8  10.47199 10.47210 10.47220 10.47228 10.47235 10.47240 10.04125 10.04116
#> 9  27.72714 27.72719 27.72723 27.72727 27.72730 27.72732 28.48634 28.48646
#> 10 29.04087 29.04078 29.04071 29.04064 29.04059 29.04055 30.22528 30.22539
#> 11 23.88624 23.88607 23.88592 23.88580 23.88570 23.88562 22.22332 22.22303
#> 12 16.24242 16.24243 16.24243 16.24244 16.24244 16.24244 15.95261 15.95262
#> 13 17.43309 17.43310 17.43310 17.43310 17.43310 17.43310 18.37759 18.37776
#> 14 13.91071 13.91069 13.91068 13.91066 13.91065 13.91065 12.81524 12.81502
#> 15 16.04630 16.04640 16.04649 16.04656 16.04662 16.04667 16.35213 16.35220
#> 16 28.05496 28.05497 28.05499 28.05500 28.05501 28.05501 28.32066 28.32071
#> 17 28.38535 28.38564 28.38588 28.38608 28.38626 28.38640 28.52069 28.52074
#> 18 29.05662 29.05670 29.05677 29.05682 29.05687 29.05691 28.58320 28.58320
#> 19 20.14256 20.14248 20.14241 20.14235 20.14230 20.14226 19.44487 19.44477
#> 20 15.08352 15.08340 15.08330 15.08322 15.08315 15.08309 15.83282 15.83292
#> 21 24.55520 24.55530 24.55539 24.55546 24.55552 24.55557 23.69870 23.69855
#>      0.1315   0.1226   0.1144   0.1066   0.0995   0.0928   0.0865   0.0807
#> 1  20.51033 20.51024 20.51016 20.69807 20.69806 20.69804 20.69803 20.69801
#> 2  20.69059 20.69056 20.69054 20.68376 20.68373 20.68370 20.68367 20.68364
#> 3  19.83847 19.83851 19.83856 20.16263 20.16272 20.16282 20.16290 20.16298
#> 4  23.40399 23.40395 23.40391 23.83173 23.83171 23.83169 23.83167 23.83164
#> 5  14.76722 14.76725 14.76728 14.79556 14.79559 14.79562 14.79565 14.79568
#> 6  16.22897 16.22906 16.22914 15.84710 15.84712 15.84713 15.84714 15.84716
#> 7  16.78608 16.78622 16.78636 16.16079 16.16079 16.16079 16.16079 16.16079
#> 8  10.04108 10.04101 10.04094 10.23419 10.23411 10.23404 10.23397 10.23390
#> 9  28.48657 28.48667 28.48677 28.37624 28.37626 28.37629 28.37631 28.37634
#> 10 30.22550 30.22560 30.22570 30.87066 30.87085 30.87103 30.87119 30.87134
#> 11 22.22276 22.22251 22.22227 21.17596 21.17560 21.17526 21.17494 21.17465
#> 12 15.95262 15.95263 15.95263 16.28712 16.28724 16.28735 16.28745 16.28754
#> 13 18.37792 18.37807 18.37821 18.41374 18.41388 18.41402 18.41415 18.41427
#> 14 12.81481 12.81462 12.81444 13.12291 13.12287 13.12283 13.12279 13.12275
#> 15 16.35227 16.35234 16.35240 16.41501 16.41508 16.41515 16.41522 16.41528
#> 16 28.32075 28.32079 28.32083 28.40416 28.40420 28.40424 28.40428 28.40432
#> 17 28.52078 28.52082 28.52085 27.41552 27.41528 27.41506 27.41486 27.41467
#> 18 28.58319 28.58319 28.58318 29.07440 29.07453 29.07465 29.07476 29.07486
#> 19 19.44468 19.44459 19.44451 19.55978 19.55975 19.55972 19.55968 19.55965
#> 20 15.83302 15.83311 15.83320 15.56660 15.56664 15.56667 15.56671 15.56675
#> 21 23.69840 23.69827 23.69814 24.00406 24.00398 24.00391 24.00384 24.00377
#>      0.0752   0.0702   0.0654   0.0610   0.0569   0.0531   0.0495   0.0462
#> 1  20.69799 20.69797 20.69795 20.69794 20.69792 20.69790 20.69788 20.69786
#> 2  20.68361 20.68359 20.68356 20.68354 20.68351 20.68349 20.68347 20.68345
#> 3  20.16306 20.16313 20.16319 20.16325 20.16331 20.16336 20.16341 20.16346
#> 4  23.83162 23.83160 23.83159 23.83157 23.83155 23.83154 23.83152 23.83151
#> 5  14.79570 14.79572 14.79575 14.79577 14.79578 14.79580 14.79582 14.79583
#> 6  15.84717 15.84719 15.84720 15.84721 15.84722 15.84724 15.84725 15.84726
#> 7  16.16080 16.16080 16.16081 16.16082 16.16083 16.16083 16.16084 16.16085
#> 8  10.23384 10.23378 10.23373 10.23368 10.23363 10.23358 10.23354 10.23350
#> 9  28.37636 28.37639 28.37641 28.37643 28.37646 28.37648 28.37650 28.37652
#> 10 30.87148 30.87161 30.87174 30.87185 30.87196 30.87206 30.87215 30.87224
#> 11 21.17437 21.17411 21.17387 21.17364 21.17343 21.17324 21.17305 21.17288
#> 12 16.28763 16.28771 16.28778 16.28785 16.28791 16.28797 16.28802 16.28807
#> 13 18.41438 18.41449 18.41459 18.41468 18.41477 18.41485 18.41493 18.41500
#> 14 13.12272 13.12269 13.12265 13.12262 13.12259 13.12256 13.12253 13.12251
#> 15 16.41534 16.41540 16.41545 16.41550 16.41555 16.41559 16.41563 16.41567
#> 16 28.40435 28.40438 28.40441 28.40444 28.40447 28.40449 28.40452 28.40454
#> 17 27.41449 27.41433 27.41419 27.41405 27.41392 27.41381 27.41370 27.41360
#> 18 29.07495 29.07504 29.07512 29.07520 29.07527 29.07534 29.07540 29.07546
#> 19 19.55962 19.55959 19.55956 19.55954 19.55951 19.55948 19.55945 19.55943
#> 20 15.56678 15.56682 15.56685 15.56688 15.56691 15.56694 15.56697 15.56700
#> 21 24.00371 24.00366 24.00360 24.00355 24.00350 24.00346 24.00341 24.00337
#>      0.0431   0.0402   0.0374   0.0349   0.0326   0.0304   0.0283   0.0264
#> 1  20.69784 20.69782 20.69781 20.69779 20.69777 20.69776 20.69774 20.69772
#> 2  20.68343 20.68341 20.68339 20.68337 20.68335 20.68334 20.68332 20.68331
#> 3  20.16350 20.16354 20.16358 20.16361 20.16364 20.16367 20.16370 20.16373
#> 4  23.83149 23.83148 23.83147 23.83145 23.83144 23.83143 23.83142 23.83141
#> 5  14.79585 14.79586 14.79587 14.79589 14.79590 14.79591 14.79592 14.79593
#> 6  15.84727 15.84728 15.84729 15.84730 15.84731 15.84732 15.84733 15.84734
#> 7  16.16086 16.16087 16.16088 16.16089 16.16090 16.16091 16.16092 16.16093
#> 8  10.23346 10.23343 10.23339 10.23336 10.23333 10.23331 10.23328 10.23326
#> 9  28.37654 28.37656 28.37657 28.37659 28.37661 28.37663 28.37664 28.37666
#> 10 30.87232 30.87240 30.87247 30.87253 30.87260 30.87265 30.87271 30.87276
#> 11 21.17272 21.17256 21.17242 21.17229 21.17217 21.17205 21.17194 21.17184
#> 12 16.28812 16.28816 16.28820 16.28823 16.28827 16.28830 16.28832 16.28835
#> 13 18.41507 18.41514 18.41520 18.41525 18.41531 18.41536 18.41540 18.41545
#> 14 13.12248 13.12246 13.12243 13.12241 13.12239 13.12237 13.12235 13.12233
#> 15 16.41571 16.41574 16.41578 16.41581 16.41584 16.41586 16.41589 16.41591
#> 16 28.40456 28.40458 28.40460 28.40462 28.40464 28.40465 28.40467 28.40468
#> 17 27.41351 27.41342 27.41335 27.41328 27.41321 27.41315 27.41309 27.41304
#> 18 29.07551 29.07556 29.07561 29.07565 29.07569 29.07573 29.07576 29.07579
#> 19 19.55941 19.55938 19.55936 19.55934 19.55932 19.55930 19.55928 19.55926
#> 20 15.56703 15.56705 15.56708 15.56710 15.56712 15.56715 15.56717 15.56719
#> 21 24.00334 24.00330 24.00327 24.00323 24.00321 24.00318 24.00315 24.00313
#>      0.0246   0.0230   0.0214   0.0200   0.0186   0.0174   0.0162   0.0151
#> 1  20.69771 20.69769 20.69768 20.69767 20.72698 20.72703 20.72708 20.72712
#> 2  20.68329 20.68328 20.68326 20.68325 20.67554 20.67558 20.67562 20.67565
#> 3  20.16375 20.16378 20.16380 20.16382 20.16210 20.16202 20.16195 20.16188
#> 4  23.83140 23.83139 23.83138 23.83138 23.81582 23.81580 23.81577 23.81575
#> 5  14.79593 14.79594 14.79595 14.79596 14.58525 14.58522 14.58519 14.58517
#> 6  15.84735 15.84736 15.84736 15.84737 15.77978 15.77979 15.77979 15.77980
#> 7  16.16094 16.16094 16.16095 16.16096 16.11485 16.11486 16.11487 16.11488
#> 8  10.23323 10.23321 10.23319 10.23318 10.38953 10.38955 10.38958 10.38960
#> 9  28.37667 28.37668 28.37670 28.37671 28.32324 28.32318 28.32312 28.32307
#> 10 30.87280 30.87285 30.87289 30.87293 30.97281 30.97261 30.97241 30.97222
#> 11 21.17174 21.17165 21.17157 21.17149 21.25522 21.25563 21.25602 21.25640
#> 12 16.28838 16.28840 16.28842 16.28844 16.30991 16.30984 16.30978 16.30971
#> 13 18.41549 18.41553 18.41557 18.41560 18.43910 18.43897 18.43884 18.43872
#> 14 13.12231 13.12229 13.12228 13.12226 13.05686 13.05691 13.05697 13.05702
#> 15 16.41594 16.41596 16.41598 16.41600 16.57498 16.57490 16.57482 16.57474
#> 16 28.40469 28.40471 28.40472 28.40473 28.39191 28.39186 28.39182 28.39177
#> 17 27.41299 27.41295 27.41291 27.41287 27.41529 27.41544 27.41559 27.41573
#> 18 29.07582 29.07585 29.07588 29.07590 29.12068 29.12058 29.12049 29.12040
#> 19 19.55924 19.55923 19.55921 19.55919 19.51797 19.51803 19.51810 19.51816
#> 20 15.56721 15.56723 15.56724 15.56726 15.61333 15.61330 15.61328 15.61325
#> 21 24.00310 24.00308 24.00306 24.00304 23.85885 23.85888 23.85892 23.85896
#>      0.0141   0.0131   0.0123   0.0114   0.0107   0.0099   0.0093   0.0087
#> 1  20.72716 20.72720 20.72724 20.72728 20.72732 20.72735 20.72738 20.72741
#> 2  20.67569 20.67572 20.67576 20.67579 20.67582 20.67585 20.67588 20.67591
#> 3  20.16181 20.16175 20.16169 20.16163 20.16157 20.16151 20.16146 20.16141
#> 4  23.81573 23.81571 23.81570 23.81568 23.81567 23.81566 23.81565 23.81564
#> 5  14.58515 14.58513 14.58512 14.58510 14.58510 14.58509 14.58509 14.58509
#> 6  15.77981 15.77982 15.77982 15.77983 15.77984 15.77984 15.77985 15.77986
#> 7  16.11490 16.11491 16.11492 16.11492 16.11493 16.11494 16.11495 16.11495
#> 8  10.38962 10.38964 10.38965 10.38967 10.38968 10.38969 10.38970 10.38971
#> 9  28.32302 28.32297 28.32292 28.32288 28.32284 28.32280 28.32276 28.32273
#> 10 30.97203 30.97186 30.97169 30.97152 30.97136 30.97121 30.97107 30.97093
#> 11 21.25676 21.25710 21.25742 21.25773 21.25803 21.25831 21.25857 21.25883
#> 12 16.30965 16.30959 16.30953 16.30947 16.30942 16.30937 16.30932 16.30927
#> 13 18.43860 18.43849 18.43838 18.43827 18.43817 18.43807 18.43798 18.43789
#> 14 13.05707 13.05712 13.05717 13.05722 13.05726 13.05731 13.05736 13.05740
#> 15 16.57466 16.57458 16.57451 16.57443 16.57436 16.57428 16.57421 16.57414
#> 16 28.39173 28.39169 28.39165 28.39161 28.39158 28.39154 28.39151 28.39148
#> 17 27.41586 27.41599 27.41612 27.41624 27.41635 27.41646 27.41657 27.41667
#> 18 29.12031 29.12022 29.12014 29.12006 29.11998 29.11990 29.11983 29.11976
#> 19 19.51822 19.51828 19.51834 19.51840 19.51845 19.51850 19.51855 19.51860
#> 20 15.61322 15.61319 15.61316 15.61313 15.61310 15.61307 15.61304 15.61301
#> 21 23.85900 23.85904 23.85909 23.85913 23.85917 23.85922 23.85927 23.85931
#>      0.0081   0.0075   0.0070   0.0065   0.0061   0.0057   0.0053
#> 1  20.72744 20.72747 20.72750 20.72753 20.72755 20.72757 20.72760
#> 2  20.67594 20.67596 20.67599 20.67601 20.67604 20.67606 20.67608
#> 3  20.16136 20.16131 20.16127 20.16122 20.16118 20.16114 20.16110
#> 4  23.81563 23.81563 23.81563 23.81562 23.81562 23.81562 23.81562
#> 5  14.58509 14.58510 14.58511 14.58512 14.58513 14.58514 14.58515
#> 6  15.77986 15.77987 15.77987 15.77988 15.77989 15.77989 15.77990
#> 7  16.11496 16.11497 16.11497 16.11498 16.11498 16.11498 16.11499
#> 8  10.38972 10.38972 10.38973 10.38973 10.38973 10.38973 10.38974
#> 9  28.32270 28.32267 28.32264 28.32262 28.32259 28.32257 28.32255
#> 10 30.97079 30.97066 30.97054 30.97042 30.97031 30.97020 30.97010
#> 11 21.25907 21.25929 21.25951 21.25971 21.25991 21.26009 21.26026
#> 12 16.30922 16.30917 16.30913 16.30909 16.30905 16.30901 16.30897
#> 13 18.43780 18.43771 18.43763 18.43756 18.43748 18.43741 18.43734
#> 14 13.05745 13.05749 13.05753 13.05757 13.05761 13.05765 13.05769
#> 15 16.57407 16.57400 16.57394 16.57387 16.57381 16.57375 16.57369
#> 16 28.39145 28.39142 28.39140 28.39137 28.39135 28.39132 28.39130
#> 17 27.41676 27.41686 27.41694 27.41703 27.41711 27.41719 27.41726
#> 18 29.11969 29.11962 29.11956 29.11950 29.11944 29.11938 29.11932
#> 19 19.51865 19.51870 19.51874 19.51878 19.51882 19.51886 19.51890
#> 20 15.61298 15.61295 15.61292 15.61290 15.61287 15.61284 15.61281
#> 21 23.85936 23.85941 23.85945 23.85950 23.85954 23.85959 23.85964
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 18.1 24.4 16.4 17.3 15.2 10.4 32.4 30.4 21.5 15.5 15.2 13.3 19.2
#> [16] 27.3 26.0 30.4 19.7 15.0 21.4
#> 
#> $X
#>               am       carb         cyl        disp        drat       gear
#>  [1,]  1.0488088  0.6971392 -0.05299989 -0.47932854  0.55467523  0.3735437
#>  [2,]  1.0488088  0.6971392 -0.05299989 -0.47932854  0.55467523  0.3735437
#>  [3,] -0.9534626 -1.0596516 -0.05299989  0.09510821 -1.47885081 -0.9338592
#>  [4,] -0.9534626 -0.4740547 -1.16599767 -0.59686713  0.18007833  0.3735437
#>  [5,] -0.9534626  0.1115423  1.05999788  0.54405262 -0.92587443 -0.9338592
#>  [6,] -0.9534626  0.1115423  1.05999788  0.54405262 -0.92587443 -0.9338592
#>  [7,] -0.9534626  0.1115423  1.05999788  0.54405262 -0.92587443 -0.9338592
#>  [8,] -0.9534626  0.6971392  1.05999788  2.27796784 -1.17560570 -0.9338592
#>  [9,]  1.0488088 -1.0596516 -1.16599767 -1.19781634  0.87575829  0.3735437
#> [10,]  1.0488088 -0.4740547 -1.16599767 -1.22432881  2.39198385  0.3735437
#> [11,] -0.9534626 -1.0596516 -1.16599767 -0.83194432  0.19791628 -0.9338592
#> [12,] -0.9534626 -0.4740547  1.05999788  0.91699463 -1.47885081 -0.9338592
#> [13,] -0.9534626 -0.4740547  1.05999788  0.79326979 -0.78317085 -0.9338592
#> [14,] -0.9534626  0.6971392  1.05999788  1.19979426  0.25143012 -0.9338592
#> [15,] -0.9534626 -0.4740547  1.05999788  1.64166867 -0.90803648 -0.9338592
#> [16,]  1.0488088 -1.0596516 -1.16599767 -1.19516509  0.87575829  0.3735437
#> [17,]  1.0488088 -0.4740547 -1.16599767 -0.83017682  1.50008646  1.6809466
#> [18,]  1.0488088 -0.4740547 -1.16599767 -1.05288153  0.32278191  1.6809466
#> [19,]  1.0488088  1.8683331 -0.05299989 -0.61189086  0.05521270  1.6809466
#> [20,]  1.0488088  3.0395271  1.05999788  0.76675732 -0.08749089  1.6809466
#> [21,]  1.0488088 -0.4740547 -1.16599767 -0.82399058  0.92927213  0.3735437
#>               hp        qsec         vs         wt
#>  [1,] -0.4586738 -0.78402341 -0.7844645 -0.5091523
#>  [2,] -0.4586738 -0.41299261 -0.7844645 -0.2246370
#>  [3,] -0.5332260  1.70718336  1.2747549  0.4280747
#>  [4,] -1.1743753  1.56142126  1.2747549  0.1268232
#>  [5,]  0.5850576 -0.16122171 -0.7844645  1.1086801
#>  [6,]  0.5850576 -0.02871072 -0.7844645  0.7293263
#>  [7,]  0.5850576  0.23631128 -0.7844645  0.7851136
#>  [8,]  0.9578188  0.22306018 -0.7844645  2.4252610
#>  [9,] -1.1147335  1.21026712  1.2747549 -0.9777659
#> [10,] -1.3234798  0.58083988  1.2747549 -1.6304776
#> [11,] -0.6525096  1.56804681  1.2747549 -0.6820930
#> [12,]  0.1377441 -0.51237586 -0.7844645  0.4950195
#> [13,]  0.1377441 -0.22747721 -0.7844645  0.4001811
#> [14,]  1.5542368 -1.47970615 -0.7844645  0.8520584
#> [15,]  0.5105054 -0.39311596 -0.7844645  0.8576371
#> [16,] -1.1147335  0.83261077  1.2747549 -1.2734387
#> [17,] -0.7419723 -0.62501021 -0.7844645 -1.0447106
#> [18,] -0.4139425 -0.49249921  1.2747549 -1.7442837
#> [19,]  0.5105054 -1.42007620 -0.7844645 -0.3417903
#> [20,]  2.8961772 -2.01637569 -0.7844645  0.5508069
#> [21,] -0.4735843  0.63384427  1.2747549 -0.3306329
#> attr(,"center")
#>  [1]   0.4761905   2.8095238   6.0952381 214.2380952   3.5890476   3.7142857
#>  [7] 140.7619048  17.6433333   0.3809524   3.0763333
#> attr(,"scale")
#>  [1]   0.4994328   1.7076592   1.7969488 113.1543214   0.5606026   0.7648752
#>  [7]  67.0670629   1.5093087   0.4856209   0.8962609
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 38
#> 
#> $lambda.min
#> [1] 0.4015295
#> 
#> $null.dev
#> [1] 34.85918
#> 
#> $Bias
#> [1] 4.360241
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 9.992771