Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 36.86470 33.11345 28.99420 25.68762 23.16137 21.29026 19.96696 19.09965
#>   [9] 18.60980 18.43042 18.50446 18.78340 19.22609 19.79771 20.46891 21.10559
#>  [17] 20.87041 20.72942 20.41295 20.06408 19.75698 19.46919 19.15939 18.79292
#>  [25] 18.24978 17.94451 17.67930 16.98784 16.91559 16.29382 18.12621 17.65703
#>  [33] 17.62456 17.66850 17.74892 18.12658 18.71125 23.92785 24.04645 24.08447
#>  [41] 24.13204 24.09370 23.59200 22.47604 24.16153 24.45049 24.08485 28.92108
#>  [49] 34.60545 36.62829 36.63077 36.72060 37.38915 43.06137 43.06715 46.46092
#>  [57] 45.17604 46.63495 46.59705 48.02311 48.02238 48.07328 48.13301 48.06946
#>  [65] 49.75592 49.75888 49.76176 49.06260 49.90915 49.91073 49.91227 49.91377
#>  [73] 49.91524 50.16177 50.16330 50.16479 50.16625 50.16767 49.84813 49.84821
#>  [81] 49.84828 49.84834 49.84840 49.84845 49.84850 49.84855 49.84858 49.84862
#>  [89] 49.84865 49.84868 49.84870 49.84872 49.84874 49.88068 49.88103 49.88138
#>  [97] 49.88173 49.88213 49.88634 49.88667
#> 
#> $cvse
#>   [1]  9.160411  8.874625  8.257264  7.641549  7.053046  6.497667  5.998656
#>   [8]  5.591708  5.318548  5.216348  5.304555  5.577073  6.006377  6.555415
#>  [15]  7.188121  7.780205  7.740308  7.712937  7.624629  7.458225  7.307214
#>  [22]  7.168958  7.031642  6.892889  6.751430  6.632516  6.527968  6.459619
#>  [29]  6.399506  6.410195  6.213561  6.188614  6.143846  6.151797  6.159833
#>  [36]  6.140362  6.187218  7.334938  7.332477  7.239782  7.276947  7.272876
#>  [43]  7.235409  7.159068  9.245690  9.235495  9.206963 12.921273 13.860137
#>  [50] 14.378460 14.379619 14.370546 14.299041 17.777010 17.777623 17.981465
#>  [57] 17.456867 19.469181 19.474011 19.454258 19.455539 19.459528 19.458632
#>  [64] 19.421280 20.229510 20.230940 20.232329 20.092536 20.611140 20.612180
#>  [71] 20.613191 20.614175 20.615131 20.631895 20.632809 20.633699 20.634564
#>  [78] 20.635405 20.457723 20.457791 20.457854 20.457910 20.457961 20.458006
#>  [85] 20.458046 20.458081 20.458111 20.458137 20.458158 20.458175 20.458188
#>  [92] 20.458197 20.458203 20.515126 20.515479 20.515828 20.516175 20.516513
#>  [99] 20.516422 20.516760
#> 
#> $fold
#>  [1]  1  7  8  5  6 10  3  8  3  9  2  1  4  4  7  2  6 10  5  1  9
#> 
#> $lambda
#>   [1] 4.973102823 4.637932337 4.325351220 4.033837025 3.761969910 3.508425729
#>   [7] 3.271969577 3.051449778 2.845792276 2.653995401 2.475124994 2.308309853
#>  [13] 2.152737495 2.007650193 1.872341289 1.746151753 1.628466969 1.518713745
#>  [19] 1.416357522 1.320899765 1.231875542 1.148851253 1.071422524 0.999212232
#>  [25] 0.931868672 0.869063842 0.810491848 0.755867410 0.704924476 0.657414926
#>  [31] 0.613107360 0.571785976 0.533249515 0.497310282 0.463793234 0.432535122
#>  [37] 0.403383702 0.376196991 0.350842573 0.327196958 0.305144978 0.284579228
#>  [43] 0.265399540 0.247512499 0.230830985 0.215273749 0.200765019 0.187234129
#>  [49] 0.174615175 0.162846697 0.151871375 0.141635752 0.132089977 0.123187554
#>  [55] 0.114885125 0.107142252 0.099921223 0.093186867 0.086906384 0.081049185
#>  [61] 0.075586741 0.070492448 0.065741493 0.061310736 0.057178598 0.053324951
#>  [67] 0.049731028 0.046379323 0.043253512 0.040338370 0.037619699 0.035084257
#>  [73] 0.032719696 0.030514498 0.028457923 0.026539954 0.024751250 0.023083099
#>  [79] 0.021527375 0.020076502 0.018723413 0.017461518 0.016284670 0.015187137
#>  [85] 0.014163575 0.013208998 0.012318755 0.011488513 0.010714225 0.009992122
#>  [91] 0.009318687 0.008690638 0.008104918 0.007558674 0.007049245 0.006574149
#>  [97] 0.006131074 0.005717860 0.005332495 0.004973103
#> 
#> $fit
#> $beta
#>               4.9731     4.6379     4.3254     4.0338    3.7620    3.5084
#> (Intercept) 21.91429 23.4843314 25.0307641 26.4729724 27.817981 29.072340
#> am           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> cyl          0.00000 -0.2794149 -0.5546275 -0.8112917 -1.050658 -1.273891
#> disp         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> wt           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#>                3.2720    3.0514    2.8458    2.6540    2.4751    2.3083
#> (Intercept) 30.242159 31.333137 32.350587 33.299463 34.184389 35.009674
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.482079 -1.676236 -1.857308 -2.026176 -2.183662 -2.330535
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                2.1527    2.0077    1.8723    1.7462    1.6285    1.5187
#> (Intercept) 35.779337 36.497127 37.166541 37.790839 38.257812 38.257812
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.467509 -2.595251 -2.714384 -2.825488 -2.908594 -2.908594
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.4164    1.3209    1.2319    1.1489     1.0714     0.9992
#> (Intercept) 38.257812 38.257812 38.257812 38.257812 37.8555129 37.4797659
#> am           0.000000  0.000000  0.000000  0.000000  0.3170835  0.6132387
#> carb         0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> cyl         -2.908594 -2.908594 -2.908594 -2.908594 -2.8719310 -2.8376880
#> disp         0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#>                 0.9319    0.8691    0.8105    0.7559    0.7049    0.6574
#> (Intercept) 37.1293426 36.802537 40.110096 40.109856 40.109705 40.109610
#> am           0.8894344  1.147016  0.000000  0.000000  0.000000  0.000000
#> carb         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.8057529 -2.775970 -1.599787 -1.599585 -1.599458 -1.599378
#> disp         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.0000000  0.000000 -3.222125 -3.222438 -3.222634 -3.222758
#>                0.6131    0.5718    0.5332    0.4973    0.4638      0.4325
#> (Intercept) 40.109550 40.109513 40.109489 40.109474 40.109465 40.04021911
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000 -0.06773905
#> cyl         -1.599328 -1.599296 -1.599276 -1.599264 -1.599256 -1.54862891
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> wt          -3.222836 -3.222885 -3.222916 -3.222936 -3.222948 -3.22828139
#>                 0.4034     0.3762     0.3508    0.3272     0.3051     0.2846
#> (Intercept) 39.9568217 39.8788370 39.8061055  6.192393  7.7475013  9.3193324
#> am           0.0000000  0.0000000  0.0000000  3.699909  3.8499222  4.0004170
#> carb        -0.1492771 -0.2255058 -0.2965992  0.000000 -0.2093335 -0.4204903
#> cyl         -1.4876500 -1.4306269 -1.3774450  0.000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  1.346475  1.2512013  1.1550034
#> vs           0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> wt          -3.2347611 -3.2408421 -3.2465137 -3.681057 -3.4484365 -3.2141371
#>                 0.2654     0.2475     0.2308     0.2153     0.2008     0.1872
#> (Intercept)  9.6945097  9.6924313  9.6909179  9.6898317  9.6890601  9.6885165
#> am           4.0347204  4.0354196  4.0359103  4.0362531  4.0364914  4.0366567
#> carb        -0.4702477 -0.4703424 -0.4704033 -0.4704428 -0.4704686 -0.4704856
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         1.1321860  1.1322332  1.1322692  1.1322959  1.1323153  1.1323292
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -3.1594226 -3.1590426 -3.1587804 -3.1585999 -3.1584757 -3.1583903
#>                 0.1746     0.1628     0.1519     0.1416       0.1321
#> (Intercept)  9.6881357  9.6878703  9.6876859  9.6875582 12.543194824
#> am           4.0367710  4.0368498  4.0369042  4.0369416  3.979096773
#> carb        -0.4704968 -0.4705043 -0.4705093 -0.4705127 -0.355767032
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.000000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.000000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.000000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.000000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000 -0.009673034
#> qsec         1.1323391  1.1323461  1.1323510  1.1323543  0.977208926
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.000000000
#> wt          -3.1583317 -3.1582915 -3.1582639 -3.1582449 -2.831699344
#>                   0.1232       0.1149      0.1071      0.0999       0.0932
#> (Intercept) 12.543076699 12.542925278 12.54276980 12.54262585 12.542500558
#> am           3.978618701  3.978257062  3.97798742  3.97778933  3.977646029
#> carb        -0.355580523 -0.355466279 -0.35540108 -0.35536823 -0.355355906
#> cyl          0.000000000  0.000000000  0.00000000  0.00000000  0.000000000
#> disp         0.000000000  0.000000000  0.00000000  0.00000000  0.000000000
#> drat         0.000000000  0.000000000  0.00000000  0.00000000  0.000000000
#> gear         0.000000000  0.000000000  0.00000000  0.00000000  0.000000000
#> hp          -0.009672483 -0.009671141 -0.00966946 -0.00966772 -0.009666081
#> qsec         0.977258272  0.977299826  0.97733376  0.97736077  0.977381822
#> vs           0.000000000  0.000000000  0.00000000  0.00000000  0.000000000
#> wt          -2.832080890 -2.832390982 -2.83263772 -2.83283039 -2.832978262
#>                  0.0869      0.0810      0.0756      0.0705      0.0657
#> (Intercept)  6.82197870  6.82052785  7.58068349  7.58032744  6.17775790
#> am           4.15420761  4.15429543  4.32876892  4.32932177  4.65045564
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.37858363
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.02281468  0.02281914  0.02395413  0.02395876  0.03320614
#> drat         0.00000000  0.00000000 -0.48844918 -0.48895417 -0.79397449
#> gear         0.61612605  0.61621898  0.78164273  0.78174763  0.74336240
#> hp          -0.02881439 -0.02881578 -0.03058667 -0.03058809 -0.04001974
#> qsec         1.14670598  1.14677047  1.16838010  1.16845368  1.32162410
#> vs           1.03080959  1.03088428  1.15928673  1.15963301  1.54152841
#> wt          -3.88505567 -3.88533136 -3.91207044 -3.91228657 -4.58298418
#>                  0.0613      0.0572      0.0533      0.0497      0.0464
#> (Intercept)  6.17678517  6.17584241  6.17492870  6.17404314  6.17318487
#> am           4.65065818  4.65085448  4.65104473  4.65122912  4.65140783
#> carb         0.37872191  0.37885592  0.37898580  0.37911168  0.37923369
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.03321143  0.03321656  0.03322153  0.03322634  0.03323101
#> drat        -0.79419121 -0.79440126 -0.79460483 -0.79480213 -0.79499335
#> gear         0.74341170  0.74345949  0.74350581  0.74355070  0.74359420
#> hp          -0.04002389 -0.04002791 -0.04003181 -0.04003559 -0.04003925
#> qsec         1.32171425  1.32180161  1.32188629  1.32196835  1.32204789
#> vs           1.54171857  1.54190287  1.54208149  1.54225460  1.54242238
#> wt          -4.58334510 -4.58369491 -4.58403394 -4.58436252 -4.58468097
#>                  0.0433      0.0403      0.0376      0.0351      0.0327
#> (Intercept)  6.17235304  6.17154684  6.17076548  6.17000819  6.16927423
#> am           4.65158103  4.65174890  4.65191159  4.65206928  4.65222210
#> carb         0.37935193  0.37946653  0.37957760  0.37968525  0.37978958
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.03323553  0.03323992  0.03324417  0.03324828  0.03325228
#> drat        -0.79517868 -0.79535830 -0.79553239 -0.79570111 -0.79586464
#> gear         0.74363637  0.74367723  0.74371684  0.74375522  0.74379243
#> hp          -0.04004280 -0.04004624 -0.04004957 -0.04005280 -0.04005593
#> qsec         1.32212498  1.32219969  1.32227210  1.32234228  1.32241030
#> vs           1.54258500  1.54274260  1.54289535  1.54304339  1.54318687
#> wt          -4.58498962 -4.58528876 -4.58557868 -4.58585966 -4.58613199
#>                  0.0305      0.0285      0.0265      0.0248      0.0231
#> (Intercept)  6.16856289  6.16787346  6.16720528  6.16655768  6.16593004
#> am           4.65237021  4.65251376  4.65265289  4.65278773  4.65291842
#> carb         0.37989070  0.37998870  0.38008368  0.38017574  0.38026496
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.03325614  0.03325989  0.03326353  0.03326705  0.03327046
#> drat        -0.79602312 -0.79617673 -0.79632560 -0.79646988 -0.79660972
#> gear         0.74382849  0.74386343  0.74389730  0.74393013  0.74396194
#> hp          -0.04005897 -0.04006191 -0.04006476 -0.04006752 -0.04007020
#> qsec         1.32247622  1.32254011  1.32260203  1.32266204  1.32272021
#> vs           1.54332593  1.54346070  1.54359132  1.54371792  1.54384062
#> wt          -4.58639593 -4.58665174 -4.58689967 -4.58713995 -4.58737284
#>                  0.0215      0.0201      0.0187      0.0175      0.0163
#> (Intercept)  6.16532173  6.16473217  6.16416077  6.16360698  6.16307024
#> am           4.65304508  4.65316784  4.65328682  4.65340213  4.65351388
#> carb         0.38035143  0.38043524  0.38051646  0.38059518  0.38067148
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.03327377  0.03327698  0.03328008  0.03328309  0.03328601
#> drat        -0.79674525 -0.79687660 -0.79700391 -0.79712729 -0.79724687
#> gear         0.74399278  0.74402266  0.74405163  0.74407970  0.74410690
#> hp          -0.04007280 -0.04007531 -0.04007775 -0.04008011 -0.04008240
#> qsec         1.32277658  1.32283122  1.32288417  1.32293549  1.32298523
#> vs           1.54395953  1.54407479  1.54418649  1.54429475  1.54439967
#> wt          -4.58759854 -4.58781730 -4.58802931 -4.58823479 -4.58843395
#>                  0.0152      0.0142      0.0132      0.0123      0.0115
#> (Intercept)  6.16255005  6.16204588  6.16155725  6.16108367  6.16062468
#> am           4.65362220  4.65372718  4.65382892  4.65392753  4.65402310
#> carb         0.38074542  0.38081709  0.38088655  0.38095387  0.38101912
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.03328884  0.03329158  0.03329424  0.03329682  0.03329931
#> drat        -0.79736277 -0.79747510 -0.79758397 -0.79768948 -0.79779174
#> gear         0.74413327  0.74415883  0.74418360  0.74420760  0.74423087
#> hp          -0.04008462 -0.04008677 -0.04008886 -0.04009088 -0.04009284
#> qsec         1.32303344  1.32308016  1.32312544  1.32316933  1.32321186
#> vs           1.54450137  1.54459992  1.54469545  1.54478802  1.54487775
#> wt          -4.58862696 -4.58881403 -4.58899533 -4.58917105 -4.58934136
#>                  0.0107      0.0100      0.0093      0.0087      0.0081
#> (Intercept)  6.16017984  6.15974870  6.15933084  6.15892586  6.15853335
#> am           4.65411572  4.65420549  4.65429250  4.65437682  4.65445855
#> carb         0.38108235  0.38114364  0.38120303  0.38126060  0.38131640
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.03330173  0.03330408  0.03330635  0.03330855  0.03331068
#> drat        -0.79789085 -0.79798691 -0.79808000 -0.79817023 -0.79825768
#> gear         0.74425342  0.74427527  0.74429645  0.74431698  0.74433687
#> hp          -0.04009473 -0.04009657 -0.04009836 -0.04010008 -0.04010176
#> qsec         1.32325309  1.32329304  1.32333177  1.32336930  1.32340567
#> vs           1.54496471  1.54504900  1.54513068  1.54520985  1.54528658
#> wt          -4.58950641 -4.58966639 -4.58982143 -4.58997169 -4.59011733
#>                  0.0076      0.0070      0.0066      0.0061      0.0057
#> (Intercept)  6.88088404  6.88152566  6.88215558  6.88277460  6.88338332
#> am           4.63611362  4.63621280  4.63630999  4.63640512  4.63649816
#> carb         0.41063269  0.41078411  0.41093210  0.41107673  0.41121809
#> cyl         -0.07617970 -0.07631262 -0.07644301 -0.07657094 -0.07669645
#> disp         0.03427350  0.03427884  0.03428406  0.03428916  0.03429414
#> drat        -0.82883897 -0.82903769 -0.82923216 -0.82942241 -0.82960845
#> gear         0.72180985  0.72180332  0.72179661  0.72178969  0.72178258
#> hp          -0.04050587 -0.04050964 -0.04051331 -0.04051689 -0.04052037
#> qsec         1.31160712  1.31164557  1.31168322  1.31172004  1.31175603
#> vs           1.54189273  1.54202169  1.54214735  1.54226976  1.54238900
#> wt          -4.61525437 -4.61554464 -4.61582856 -4.61610620 -4.61637763
#>                  0.0053      0.0050
#> (Intercept)  6.88398220  6.88457159
#> am           4.63658907  4.63667785
#> carb         0.41135624  0.41149127
#> cyl         -0.07681959 -0.07694040
#> disp         0.03429902  0.03430378
#> drat        -0.82979033 -0.82996810
#> gear         0.72177527  0.72176777
#> hp          -0.04052378 -0.04052709
#> qsec         1.31179117  1.31182547
#> vs           1.54250513  1.54261821
#> wt          -4.61664292 -4.61690217
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   6   5   5   5  35   1   1   1   1   1   1   1   1  15
#>  [37]  14  14  14  63  47  47  11   1   1   1   1   1   1   1   1   1  28   1
#>  [55]   1   1   1   1 143   1 110   1 146   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1  63   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.973102823 4.637932337 4.325351220 4.033837025 3.761969910 3.508425729
#>   [7] 3.271969577 3.051449778 2.845792276 2.653995401 2.475124994 2.308309853
#>  [13] 2.152737495 2.007650193 1.872341289 1.746151753 1.628466969 1.518713745
#>  [19] 1.416357522 1.320899765 1.231875542 1.148851253 1.071422524 0.999212232
#>  [25] 0.931868672 0.869063842 0.810491848 0.755867410 0.704924476 0.657414926
#>  [31] 0.613107360 0.571785976 0.533249515 0.497310282 0.463793234 0.432535122
#>  [37] 0.403383702 0.376196991 0.350842573 0.327196958 0.305144978 0.284579228
#>  [43] 0.265399540 0.247512499 0.230830985 0.215273749 0.200765019 0.187234129
#>  [49] 0.174615175 0.162846697 0.151871375 0.141635752 0.132089977 0.123187554
#>  [55] 0.114885125 0.107142252 0.099921223 0.093186867 0.086906384 0.081049185
#>  [61] 0.075586741 0.070492448 0.065741493 0.061310736 0.057178598 0.053324951
#>  [67] 0.049731028 0.046379323 0.043253512 0.040338370 0.037619699 0.035084257
#>  [73] 0.032719696 0.030514498 0.028457923 0.026539954 0.024751250 0.023083099
#>  [79] 0.021527375 0.020076502 0.018723413 0.017461518 0.016284670 0.015187137
#>  [85] 0.014163575 0.013208998 0.012318755 0.011488513 0.010714225 0.009992122
#>  [91] 0.009318687 0.008690638 0.008104918 0.007558674 0.007049245 0.006574149
#>  [97] 0.006131074 0.005717860 0.005332495 0.004973103
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 26
#> 
#> $loss
#>   [1] 733.8457 639.5319 555.9403 486.3034 428.5968 381.0738 342.2283 310.7623
#>   [9] 285.5583 265.6549 250.2256 238.5609 230.0522 224.1781 220.4926 218.6146
#>  [17] 218.1936 218.1936 218.1936 218.1936 218.1936 218.1936 211.1799 205.3806
#>  [25] 200.6262 196.7611 156.7662 156.7662 156.7662 156.7662 156.7662 156.7662
#>  [33] 156.7662 156.7662 156.7662 154.6087 152.4462 150.8538 149.7427 125.1942
#>  [41] 121.2165 119.5099 119.4450 119.4450 119.4450 119.4449 119.4449 119.4449
#>  [49] 119.4449 119.4449 119.4449 119.4449 117.4017 117.4017 117.4017 117.4017
#>  [57] 117.4017 117.4017 110.7125 110.7125 110.3339 110.3339 109.3480 109.3480
#>  [65] 109.3480 109.3479 109.3479 109.3479 109.3479 109.3479 109.3479 109.3479
#>  [73] 109.3479 109.3479 109.3479 109.3479 109.3479 109.3478 109.3478 109.3478
#>  [81] 109.3478 109.3478 109.3478 109.3478 109.3478 109.3478 109.3478 109.3478
#>  [89] 109.3478 109.3478 109.3478 109.3478 109.3478 109.3242 109.3242 109.3242
#>  [97] 109.3242 109.3242 109.3242 109.3242
#> 
#> $linear.predictors
#>      4.9731   4.6379   4.3254   4.0338   3.7620   3.5084   3.2720   3.0514
#> 1  21.91429 21.80784 21.70300 21.60522 21.51404 21.42899 21.34968 21.27572
#> 2  21.91429 21.80784 21.70300 21.60522 21.51404 21.42899 21.34968 21.27572
#> 3  21.91429 22.36667 22.81225 23.22781 23.61535 23.97678 24.31384 24.62819
#> 4  21.91429 21.24901 20.59374 19.98264 19.41272 18.88121 18.38553 17.92325
#> 5  21.91429 22.36667 22.81225 23.22781 23.61535 23.97678 24.31384 24.62819
#> 6  21.91429 21.80784 21.70300 21.60522 21.51404 21.42899 21.34968 21.27572
#> 7  21.91429 21.80784 21.70300 21.60522 21.51404 21.42899 21.34968 21.27572
#> 8  21.91429 21.24901 20.59374 19.98264 19.41272 18.88121 18.38553 17.92325
#> 9  21.91429 21.24901 20.59374 19.98264 19.41272 18.88121 18.38553 17.92325
#> 10 21.91429 22.36667 22.81225 23.22781 23.61535 23.97678 24.31384 24.62819
#> 11 21.91429 22.36667 22.81225 23.22781 23.61535 23.97678 24.31384 24.62819
#> 12 21.91429 22.36667 22.81225 23.22781 23.61535 23.97678 24.31384 24.62819
#> 13 21.91429 22.36667 22.81225 23.22781 23.61535 23.97678 24.31384 24.62819
#> 14 21.91429 21.24901 20.59374 19.98264 19.41272 18.88121 18.38553 17.92325
#> 15 21.91429 22.36667 22.81225 23.22781 23.61535 23.97678 24.31384 24.62819
#> 16 21.91429 22.36667 22.81225 23.22781 23.61535 23.97678 24.31384 24.62819
#> 17 21.91429 22.36667 22.81225 23.22781 23.61535 23.97678 24.31384 24.62819
#> 18 21.91429 21.24901 20.59374 19.98264 19.41272 18.88121 18.38553 17.92325
#> 19 21.91429 21.80784 21.70300 21.60522 21.51404 21.42899 21.34968 21.27572
#> 20 21.91429 21.24901 20.59374 19.98264 19.41272 18.88121 18.38553 17.92325
#> 21 21.91429 22.36667 22.81225 23.22781 23.61535 23.97678 24.31384 24.62819
#>      2.8458   2.6540   2.4751   2.3083   2.1527   2.0077   1.8723   1.7462
#> 1  21.20674 21.14241 21.08241 21.02646 20.97428 20.92562 20.88023 20.83791
#> 2  21.20674 21.14241 21.08241 21.02646 20.97428 20.92562 20.88023 20.83791
#> 3  24.92136 25.19476 25.44974 25.68753 25.90930 26.11612 26.30900 26.48889
#> 4  17.49212 17.09006 16.71509 16.36539 16.03926 15.73512 15.45147 15.18693
#> 5  24.92136 25.19476 25.44974 25.68753 25.90930 26.11612 26.30900 26.48889
#> 6  21.20674 21.14241 21.08241 21.02646 20.97428 20.92562 20.88023 20.83791
#> 7  21.20674 21.14241 21.08241 21.02646 20.97428 20.92562 20.88023 20.83791
#> 8  17.49212 17.09006 16.71509 16.36539 16.03926 15.73512 15.45147 15.18693
#> 9  17.49212 17.09006 16.71509 16.36539 16.03926 15.73512 15.45147 15.18693
#> 10 24.92136 25.19476 25.44974 25.68753 25.90930 26.11612 26.30900 26.48889
#> 11 24.92136 25.19476 25.44974 25.68753 25.90930 26.11612 26.30900 26.48889
#> 12 24.92136 25.19476 25.44974 25.68753 25.90930 26.11612 26.30900 26.48889
#> 13 24.92136 25.19476 25.44974 25.68753 25.90930 26.11612 26.30900 26.48889
#> 14 17.49212 17.09006 16.71509 16.36539 16.03926 15.73512 15.45147 15.18693
#> 15 24.92136 25.19476 25.44974 25.68753 25.90930 26.11612 26.30900 26.48889
#> 16 24.92136 25.19476 25.44974 25.68753 25.90930 26.11612 26.30900 26.48889
#> 17 24.92136 25.19476 25.44974 25.68753 25.90930 26.11612 26.30900 26.48889
#> 18 17.49212 17.09006 16.71509 16.36539 16.03926 15.73512 15.45147 15.18693
#> 19 21.20674 21.14241 21.08241 21.02646 20.97428 20.92562 20.88023 20.83791
#> 20 17.49212 17.09006 16.71509 16.36539 16.03926 15.73512 15.45147 15.18693
#> 21 24.92136 25.19476 25.44974 25.68753 25.90930 26.11612 26.30900 26.48889
#>      1.6285   1.5187   1.4164   1.3209   1.2319   1.1489   1.0714   0.9992
#> 1  20.80625 20.80625 20.80625 20.80625 20.80625 20.80625 20.94101 21.06688
#> 2  20.80625 20.80625 20.80625 20.80625 20.80625 20.80625 20.94101 21.06688
#> 3  26.62344 26.62344 26.62344 26.62344 26.62344 26.62344 26.68487 26.74225
#> 4  14.98906 14.98906 14.98906 14.98906 14.98906 14.98906 14.88007 14.77826
#> 5  26.62344 26.62344 26.62344 26.62344 26.62344 26.62344 26.36779 26.12901
#> 6  20.80625 20.80625 20.80625 20.80625 20.80625 20.80625 20.62393 20.45364
#> 7  20.80625 20.80625 20.80625 20.80625 20.80625 20.80625 20.62393 20.45364
#> 8  14.98906 14.98906 14.98906 14.98906 14.98906 14.98906 14.88007 14.77826
#> 9  14.98906 14.98906 14.98906 14.98906 14.98906 14.98906 14.88007 14.77826
#> 10 26.62344 26.62344 26.62344 26.62344 26.62344 26.62344 26.68487 26.74225
#> 11 26.62344 26.62344 26.62344 26.62344 26.62344 26.62344 26.68487 26.74225
#> 12 26.62344 26.62344 26.62344 26.62344 26.62344 26.62344 26.68487 26.74225
#> 13 26.62344 26.62344 26.62344 26.62344 26.62344 26.62344 26.36779 26.12901
#> 14 14.98906 14.98906 14.98906 14.98906 14.98906 14.98906 14.88007 14.77826
#> 15 26.62344 26.62344 26.62344 26.62344 26.62344 26.62344 26.68487 26.74225
#> 16 26.62344 26.62344 26.62344 26.62344 26.62344 26.62344 26.68487 26.74225
#> 17 26.62344 26.62344 26.62344 26.62344 26.62344 26.62344 26.68487 26.74225
#> 18 14.98906 14.98906 14.98906 14.98906 14.98906 14.98906 15.19715 15.39150
#> 19 20.80625 20.80625 20.80625 20.80625 20.80625 20.80625 20.94101 21.06688
#> 20 14.98906 14.98906 14.98906 14.98906 14.98906 14.98906 15.19715 15.39150
#> 21 26.62344 26.62344 26.62344 26.62344 26.62344 26.62344 26.68487 26.74225
#>      0.9319   0.8691   0.8105   0.7559   0.7049   0.6574   0.6131   0.5718
#> 1  21.18426 21.29373 22.06941 22.06956 22.06965 22.06971 22.06975 22.06977
#> 2  21.18426 21.29373 21.24777 21.24784 21.24788 21.24791 21.24793 21.24794
#> 3  26.79577 26.84567 26.23562 26.23546 26.23536 26.23530 26.23526 26.23523
#> 4  14.68332 14.59478 15.80882 15.80907 15.80923 15.80934 15.80940 15.80944
#> 5  25.90633 25.69866 23.56125 23.56084 23.56057 23.56041 23.56030 23.56024
#> 6  20.29483 20.14672 19.42726 19.42716 19.42709 19.42705 19.42703 19.42701
#> 7  20.29483 20.14672 19.42726 19.42716 19.42709 19.42705 19.42703 19.42701
#> 8  14.68332 14.59478 15.29328 15.29348 15.29361 15.29369 15.29375 15.29378
#> 9  14.68332 14.59478 10.08954 10.08925 10.08906 10.08894 10.08887 10.08882
#> 10 26.79577 26.84567 26.62227 26.62215 26.62208 26.62203 26.62200 26.62198
#> 11 26.79577 26.84567 28.50722 28.50728 28.50732 28.50734 28.50736 28.50737
#> 12 26.79577 26.84567 27.79835 27.79834 27.79834 27.79834 27.79833 27.79833
#> 13 25.90633 25.69866 25.76841 25.76821 25.76808 25.76800 25.76795 25.76791
#> 14 14.68332 14.59478 15.96992 15.97019 15.97037 15.97047 15.97054 15.97059
#> 15 26.79577 26.84567 27.47614 27.47610 27.47607 27.47606 27.47605 27.47604
#> 16 26.79577 26.84567 26.81560 26.81550 26.81543 26.81539 26.81537 26.81535
#> 17 26.79577 26.84567 28.83587 28.83597 28.83603 28.83606 28.83609 28.83610
#> 18 15.57275 15.74179 17.09767 17.09805 17.09829 17.09844 17.09853 17.09859
#> 19 21.18426 21.29373 21.58609 21.58619 21.58626 21.58630 21.58633 21.58634
#> 20 15.57275 15.74179 15.80882 15.80907 15.80923 15.80934 15.80940 15.80944
#> 21 26.79577 26.84567 24.75344 24.75314 24.75295 24.75283 24.75275 24.75271
#>      0.5332   0.4973   0.4638   0.4325   0.4034   0.3762   0.3508    0.3272
#> 1  22.06979 22.06980 22.06980 22.01939 21.95874 21.90205 21.84917 22.410914
#> 2  21.24795 21.24795 21.24795 21.19618 21.13387 21.07563 21.02131 22.226270
#> 3  26.23522 26.23521 26.23520 26.28835 26.35230 26.41207 26.46781 26.410152
#> 4  15.80947 15.80948 15.80949 15.85527 15.91042 15.96199 16.01010 14.379186
#> 5  23.56020 23.56017 23.56015 23.54114 23.51817 23.49667 23.47661 25.431344
#> 6  19.42700 19.42699 19.42699 19.37220 19.30623 19.24456 19.18703 18.170052
#> 7  19.42700 19.42699 19.42699 19.37220 19.30623 19.24456 19.18703 18.977937
#> 8  15.29380 15.29381 15.29382 15.40648 15.54213 15.66896 15.78725 16.160013
#> 9  10.08879 10.08877 10.08876 10.12507 10.16871 10.20950 10.24753  9.972741
#> 10 26.62197 26.62196 26.62195 26.67575 26.74047 26.80097 26.85740 28.009847
#> 11 28.50737 28.50738 28.50738 28.49655 28.48353 28.47136 28.46001 28.884114
#> 12 27.79833 27.79833 27.79833 27.85407 27.92116 27.98388 28.04237 29.932417
#> 13 25.76789 25.76788 25.76787 25.82025 25.88326 25.94215 25.99707 24.061554
#> 14 15.97061 15.97063 15.97064 16.15216 16.37071 16.57505 16.76562 15.950108
#> 15 27.47604 27.47604 27.47604 27.53124 27.59768 27.65979 27.71772 28.217836
#> 16 26.81534 26.81534 26.81533 26.80170 26.78528 26.76992 26.75559 24.500975
#> 17 28.83611 28.83612 28.83612 28.82584 28.81347 28.80192 28.79115 27.078292
#> 18 17.09863 17.09866 17.09867 17.14658 17.20432 17.25833 17.30870 17.747241
#> 19 21.58635 21.58636 21.58636 21.39967 21.17497 20.96491 20.76900 20.566139
#> 20 15.80947 15.80948 15.80949 15.58431 15.31331 15.05997 14.82370 16.409466
#> 21 24.75268 24.75266 24.75265 24.73560 24.71503 24.69578 24.67782 24.703401
#>      0.3051   0.2846   0.2654   0.2475   0.2308   0.2153   0.2008   0.1872
#> 1  22.31996 22.22810 22.20633 22.20635 22.20636 22.20637 22.20638 22.20639
#> 2  22.14128 22.05530 22.03471 22.03484 22.03494 22.03502 22.03507 22.03510
#> 3  26.67257 26.93707 26.99910 26.99939 26.99958 26.99972 26.99981 26.99987
#> 4  14.41828 14.45816 14.46821 14.46785 14.46760 14.46743 14.46730 14.46722
#> 5  25.11877 24.80340 24.72889 24.72890 24.72892 24.72893 24.72895 24.72896
#> 6  17.94453 17.71730 17.66411 17.66382 17.66363 17.66349 17.66340 17.66334
#> 7  18.69525 18.41030 18.34342 18.34316 18.34299 18.34287 18.34279 18.34273
#> 8  16.27797 16.39719 16.42559 16.42548 16.42540 16.42533 16.42529 16.42526
#> 9  10.27420 10.57797 10.64909 10.64948 10.64975 10.64994 10.65007 10.65015
#> 10 28.16242 28.31607 28.35191 28.35220 28.35239 28.35252 28.35262 28.35268
#> 11 28.78178 28.67860 28.65435 28.65427 28.65422 28.65418 28.65416 28.65414
#> 12 29.95911 29.98589 29.99194 29.99211 29.99222 29.99230 29.99235 29.99239
#> 13 24.07431 24.08761 24.09133 24.09104 24.09083 24.09068 24.09058 24.09051
#> 14 16.29810 16.64950 16.73282 16.73269 16.73259 16.73251 16.73245 16.73241
#> 15 28.36307 28.50947 28.54382 28.54397 28.54407 28.54414 28.54419 28.54422
#> 16 24.69416 24.88907 24.93508 24.93511 24.93513 24.93514 24.93514 24.93515
#> 17 27.10657 27.13534 27.14247 27.14228 27.14214 27.14204 27.14197 27.14192
#> 18 17.97096 18.19652 18.24957 18.24970 18.24978 18.24984 18.24988 18.24991
#> 19 20.18287 19.79620 19.70503 19.70486 19.70476 19.70469 19.70465 19.70462
#> 20 15.87938 15.34441 15.21803 15.21793 15.21789 15.21786 15.21785 15.21784
#> 21 24.86445 25.02653 25.06420 25.06457 25.06482 25.06500 25.06512 25.06520
#>      0.1746   0.1628   0.1519   0.1416   0.1321   0.1232   0.1149   0.1071
#> 1  22.20639 22.20640 22.20640 22.20640 22.70100 22.70102 22.70098 22.70092
#> 2  22.03513 22.03515 22.03516 22.03517 22.52615 22.52610 22.52601 22.52590
#> 3  26.99991 26.99994 26.99996 26.99997 26.88325 26.88292 26.88270 26.88256
#> 4  14.46716 14.46711 14.46709 14.46707 14.12006 14.12024 14.12042 14.12060
#> 5  24.72896 24.72897 24.72897 24.72898 24.37095 24.37119 24.37137 24.37150
#> 6  17.66329 17.66326 17.66324 17.66323 18.07222 18.07251 18.07267 18.07276
#> 7  18.34270 18.34267 18.34265 18.34264 18.65855 18.65886 18.65905 18.65916
#> 8  16.42524 16.42522 16.42521 16.42520 16.37139 16.37137 16.37138 16.37140
#> 9  10.65021 10.65025 10.65028 10.65030 10.78288 10.78245 10.78213 10.78190
#> 10 28.35272 28.35275 28.35277 28.35279 28.32462 28.32437 28.32419 28.32406
#> 11 28.65413 28.65412 28.65411 28.65411 28.83247 28.83258 28.83263 28.83265
#> 12 29.99242 29.99244 29.99245 29.99246 29.78807 29.78797 29.78792 29.78789
#> 13 24.09046 24.09042 24.09040 24.09038 23.82296 23.82312 23.82328 23.82343
#> 14 16.73238 16.73235 16.73234 16.73233 16.89864 16.89847 16.89835 16.89828
#> 15 28.54425 28.54426 28.54427 28.54428 28.51801 28.51784 28.51771 28.51763
#> 16 24.93515 24.93515 24.93515 24.93515 25.19006 25.18990 25.18977 25.18966
#> 17 27.14189 27.14187 27.14185 27.14184 26.94817 26.94827 26.94837 26.94847
#> 18 18.24993 18.24994 18.24995 18.24995 17.73858 17.73839 17.73830 17.73829
#> 19 19.70460 19.70459 19.70458 19.70458 19.99784 19.99817 19.99836 19.99846
#> 20 15.21784 15.21784 15.21784 15.21784 14.59377 14.59421 14.59456 14.59484
#> 21 25.06526 25.06530 25.06533 25.06534 25.06036 25.06005 25.05983 25.05966
#>      0.0999   0.0932   0.0869   0.0810   0.0756   0.0705   0.0657   0.0613
#> 1  22.70084 22.70076 22.61739 22.61730 22.58111 22.58099 22.87682 22.87689
#> 2  22.52579 22.52568 22.26886 22.26873 22.23783 22.23769 22.44827 22.44829
#> 3  26.88247 26.88241 26.58262 26.58261 26.72482 26.72507 26.49230 26.49225
#> 4  14.12075 14.12088 14.11829 14.11842 14.02849 14.02854 14.09617 14.09623
#> 5  24.37160 24.37167 24.81387 24.81397 24.85171 24.85166 25.03986 25.03998
#> 6  18.07279 18.07279 18.21699 18.21679 18.12821 18.12784 18.15787 18.15777
#> 7  18.65921 18.65922 18.90501 18.90486 18.82923 18.82891 18.95084 18.95080
#> 8  16.37142 16.37145 15.46682 15.46674 15.49849 15.49841 15.22685 15.22673
#> 9  10.78172 10.78160 11.29150 11.29162 11.29597 11.29613 11.29045 11.29049
#> 10 28.32396 28.32390 28.14451 28.14450 28.21071 28.21084 28.10384 28.10381
#> 11 28.83266 28.83265 29.66285 29.66295 29.33048 29.33025 29.69370 29.69380
#> 12 29.78788 29.78788 29.91106 29.91115 29.92117 29.92131 30.02142 30.02150
#> 13 23.82355 23.82365 23.01514 23.01505 22.92365 22.92345 22.64504 22.64489
#> 14 16.89824 16.89822 17.27280 17.27295 17.54699 17.54729 17.69393 17.69409
#> 15 28.51757 28.51753 28.52727 28.52730 28.58862 28.58877 28.57496 28.57498
#> 16 25.18958 25.18952 25.01528 25.01528 24.89225 24.89193 24.40132 24.40111
#> 17 26.94855 26.94862 27.50252 27.50263 27.78390 27.78426 27.88750 27.88764
#> 18 17.73832 17.73837 18.76938 18.76974 18.62968 18.62990 18.43442 18.43446
#> 19 19.99851 19.99852 19.33477 19.33451 19.44362 19.44348 19.54408 19.54404
#> 20 14.59505 14.59520 14.14347 14.14341 14.14447 14.14463 14.28592 14.28596
#> 21 25.05954 25.05945 24.61958 24.61949 24.60860 24.60866 24.33443 24.33429
#>      0.0572   0.0533   0.0497   0.0464   0.0433   0.0403   0.0376   0.0351
#> 1  22.87695 22.87701 22.87706 22.87712 22.87717 22.87722 22.87727 22.87732
#> 2  22.44831 22.44833 22.44835 22.44837 22.44839 22.44841 22.44842 22.44844
#> 3  26.49220 26.49215 26.49210 26.49205 26.49201 26.49196 26.49192 26.49188
#> 4  14.09629 14.09634 14.09640 14.09645 14.09650 14.09655 14.09660 14.09664
#> 5  25.04010 25.04021 25.04032 25.04043 25.04053 25.04063 25.04072 25.04082
#> 6  18.15768 18.15758 18.15749 18.15741 18.15732 18.15724 18.15717 18.15709
#> 7  18.95076 18.95072 18.95068 18.95064 18.95060 18.95056 18.95053 18.95049
#> 8  15.22661 15.22649 15.22638 15.22627 15.22616 15.22606 15.22596 15.22587
#> 9  11.29053 11.29057 11.29061 11.29065 11.29068 11.29072 11.29075 11.29078
#> 10 28.10379 28.10376 28.10374 28.10372 28.10370 28.10367 28.10365 28.10363
#> 11 29.69389 29.69398 29.69407 29.69416 29.69424 29.69432 29.69440 29.69447
#> 12 30.02158 30.02165 30.02172 30.02179 30.02186 30.02192 30.02199 30.02205
#> 13 22.64474 22.64460 22.64446 22.64432 22.64420 22.64407 22.64395 22.64383
#> 14 17.69425 17.69440 17.69455 17.69470 17.69484 17.69497 17.69510 17.69523
#> 15 28.57500 28.57502 28.57504 28.57506 28.57508 28.57509 28.57511 28.57513
#> 16 24.40090 24.40070 24.40050 24.40031 24.40013 24.39995 24.39978 24.39961
#> 17 27.88777 27.88790 27.88803 27.88815 27.88827 27.88838 27.88849 27.88860
#> 18 18.43450 18.43454 18.43457 18.43460 18.43464 18.43467 18.43470 18.43473
#> 19 19.54400 19.54397 19.54393 19.54389 19.54386 19.54383 19.54379 19.54376
#> 20 14.28601 14.28605 14.28609 14.28613 14.28617 14.28621 14.28624 14.28628
#> 21 24.33416 24.33403 24.33391 24.33379 24.33367 24.33356 24.33345 24.33334
#>      0.0327   0.0305   0.0285   0.0265   0.0248   0.0231   0.0215   0.0201
#> 1  22.87737 22.87742 22.87746 22.87751 22.87755 22.87759 22.87763 22.87767
#> 2  22.44846 22.44847 22.44849 22.44850 22.44852 22.44853 22.44854 22.44856
#> 3  26.49184 26.49180 26.49176 26.49172 26.49169 26.49165 26.49162 26.49159
#> 4  14.09669 14.09673 14.09677 14.09681 14.09685 14.09689 14.09693 14.09697
#> 5  25.04091 25.04099 25.04108 25.04116 25.04124 25.04132 25.04139 25.04147
#> 6  18.15702 18.15694 18.15688 18.15681 18.15674 18.15668 18.15662 18.15656
#> 7  18.95046 18.95043 18.95040 18.95037 18.95034 18.95031 18.95028 18.95026
#> 8  15.22577 15.22568 15.22560 15.22551 15.22543 15.22535 15.22527 15.22520
#> 9  11.29082 11.29085 11.29088 11.29091 11.29093 11.29096 11.29099 11.29101
#> 10 28.10362 28.10360 28.10358 28.10356 28.10354 28.10353 28.10351 28.10350
#> 11 29.69454 29.69461 29.69468 29.69475 29.69481 29.69488 29.69494 29.69499
#> 12 30.02211 30.02217 30.02222 30.02228 30.02233 30.02238 30.02243 30.02248
#> 13 22.64372 22.64361 22.64350 22.64339 22.64329 22.64320 22.64310 22.64301
#> 14 17.69536 17.69547 17.69559 17.69570 17.69581 17.69592 17.69602 17.69612
#> 15 28.57514 28.57516 28.57517 28.57519 28.57520 28.57521 28.57523 28.57524
#> 16 24.39945 24.39929 24.39914 24.39899 24.39884 24.39870 24.39857 24.39844
#> 17 27.88870 27.88880 27.88890 27.88899 27.88909 27.88917 27.88926 27.88934
#> 18 18.43476 18.43479 18.43482 18.43484 18.43487 18.43489 18.43492 18.43494
#> 19 19.54373 19.54370 19.54367 19.54365 19.54362 19.54359 19.54357 19.54354
#> 20 14.28631 14.28634 14.28638 14.28641 14.28644 14.28646 14.28649 14.28652
#> 21 24.33324 24.33314 24.33305 24.33295 24.33286 24.33277 24.33269 24.33261
#>      0.0187   0.0175   0.0163   0.0152   0.0142   0.0132   0.0123   0.0115
#> 1  22.87770 22.87774 22.87777 22.87781 22.87784 22.87787 22.87790 22.87793
#> 2  22.44857 22.44858 22.44859 22.44861 22.44862 22.44863 22.44864 22.44865
#> 3  26.49156 26.49153 26.49150 26.49147 26.49144 26.49141 26.49139 26.49136
#> 4  14.09700 14.09703 14.09707 14.09710 14.09713 14.09716 14.09719 14.09722
#> 5  25.04154 25.04160 25.04167 25.04173 25.04180 25.04186 25.04191 25.04197
#> 6  18.15650 18.15645 18.15639 18.15634 18.15629 18.15624 18.15619 18.15615
#> 7  18.95023 18.95021 18.95018 18.95016 18.95014 18.95012 18.95009 18.95007
#> 8  15.22513 15.22506 15.22499 15.22492 15.22486 15.22480 15.22474 15.22468
#> 9  11.29104 11.29106 11.29109 11.29111 11.29113 11.29115 11.29117 11.29119
#> 10 28.10348 28.10347 28.10345 28.10344 28.10343 28.10341 28.10340 28.10339
#> 11 29.69505 29.69511 29.69516 29.69521 29.69526 29.69531 29.69536 29.69540
#> 12 30.02252 30.02257 30.02261 30.02265 30.02269 30.02273 30.02277 30.02281
#> 13 22.64292 22.64283 22.64275 22.64267 22.64259 22.64251 22.64244 22.64237
#> 14 17.69621 17.69631 17.69640 17.69648 17.69657 17.69665 17.69673 17.69681
#> 15 28.57525 28.57526 28.57527 28.57528 28.57529 28.57531 28.57532 28.57532
#> 16 24.39831 24.39819 24.39807 24.39796 24.39784 24.39774 24.39763 24.39753
#> 17 27.88942 27.88950 27.88958 27.88965 27.88972 27.88979 27.88986 27.88992
#> 18 18.43497 18.43499 18.43501 18.43503 18.43505 18.43507 18.43509 18.43511
#> 19 19.54352 19.54350 19.54347 19.54345 19.54343 19.54341 19.54339 19.54337
#> 20 14.28655 14.28657 14.28660 14.28662 14.28664 14.28667 14.28669 14.28671
#> 21 24.33253 24.33245 24.33237 24.33230 24.33223 24.33216 24.33210 24.33203
#>      0.0107   0.0100   0.0093   0.0087   0.0081   0.0076   0.0070   0.0066
#> 1  22.87796 22.87799 22.87801 22.87804 22.87807 22.88242 22.88248 22.88254
#> 2  22.44866 22.44867 22.44868 22.44868 22.44869 22.44003 22.44004 22.44004
#> 3  26.49134 26.49131 26.49129 26.49127 26.49125 26.49712 26.49709 26.49706
#> 4  14.09724 14.09727 14.09730 14.09732 14.09734 14.13275 14.13285 14.13295
#> 5  25.04203 25.04208 25.04213 25.04218 25.04223 25.05292 25.05301 25.05311
#> 6  18.15610 18.15606 18.15602 18.15598 18.15594 18.13437 18.13428 18.13419
#> 7  18.95005 18.95004 18.95002 18.95000 18.94998 18.92134 18.92127 18.92120
#> 8  15.22462 15.22457 15.22452 15.22446 15.22441 15.15520 15.15499 15.15479
#> 9  11.29121 11.29123 11.29125 11.29127 11.29128 11.34591 11.34603 11.34615
#> 10 28.10338 28.10337 28.10336 28.10335 28.10334 28.07775 28.07768 28.07762
#> 11 29.69545 29.69549 29.69553 29.69557 29.69561 29.70203 29.70211 29.70220
#> 12 30.02284 30.02288 30.02291 30.02295 30.02298 29.99030 29.99029 29.99028
#> 13 22.64230 22.64223 22.64217 22.64211 22.64204 22.68325 22.68322 22.68318
#> 14 17.69688 17.69696 17.69703 17.69709 17.69716 17.67476 17.67483 17.67490
#> 15 28.57533 28.57534 28.57535 28.57536 28.57537 28.56346 28.56345 28.56344
#> 16 24.39743 24.39734 24.39724 24.39715 24.39707 24.43510 24.43502 24.43494
#> 17 27.88999 27.89005 27.89011 27.89016 27.89022 27.92529 27.92544 27.92559
#> 18 18.43512 18.43514 18.43516 18.43517 18.43519 18.38584 18.38576 18.38569
#> 19 19.54335 19.54334 19.54332 19.54330 19.54329 19.55915 19.55916 19.55917
#> 20 14.28673 14.28675 14.28677 14.28679 14.28681 14.28744 14.28747 14.28749
#> 21 24.33197 24.33191 24.33185 24.33179 24.33174 24.35359 24.35353 24.35347
#>      0.0061   0.0057   0.0053   0.0050
#> 1  22.88259 22.88265 22.88270 22.88276
#> 2  22.44005 22.44006 22.44006 22.44007
#> 3  26.49703 26.49700 26.49697 26.49694
#> 4  14.13304 14.13314 14.13323 14.13332
#> 5  25.05320 25.05329 25.05338 25.05347
#> 6  18.13411 18.13402 18.13394 18.13385
#> 7  18.92114 18.92107 18.92101 18.92095
#> 8  15.15459 15.15439 15.15420 15.15401
#> 9  11.34627 11.34639 11.34650 11.34661
#> 10 28.07756 28.07749 28.07743 28.07738
#> 11 29.70228 29.70236 29.70244 29.70252
#> 12 29.99027 29.99027 29.99026 29.99025
#> 13 22.68315 22.68312 22.68309 22.68306
#> 14 17.67497 17.67503 17.67510 17.67516
#> 15 28.56343 28.56342 28.56341 28.56340
#> 16 24.43486 24.43479 24.43472 24.43464
#> 17 27.92574 27.92588 27.92602 27.92615
#> 18 18.38563 18.38556 18.38549 18.38543
#> 19 19.55917 19.55918 19.55919 19.55919
#> 20 14.28751 14.28754 14.28756 14.28758
#> 21 24.35342 24.35336 24.35331 24.35326
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 22.8 14.3 22.8 19.2 17.8 17.3 14.7 32.4 30.4 33.9 21.5 15.5 27.3
#> [16] 26.0 30.4 15.8 19.7 15.0 21.4
#> 
#> $X
#>               am        carb        cyl       disp        drat      gear
#>  [1,]  0.7844645  0.58671882  0.2236068 -0.2199666  0.15255079  0.000000
#>  [2,]  0.7844645  0.58671882  0.2236068 -0.2199666  0.15255079  0.000000
#>  [3,]  0.7844645 -1.09343053 -0.9503289 -0.7044433  0.04717031  0.000000
#>  [4,] -1.2747549  0.58671882  1.3975425  1.6434053 -1.30169981 -1.449138
#>  [5,] -1.2747549 -0.53338075 -0.9503289 -0.3988503  0.19470298  0.000000
#>  [6,] -1.2747549  0.58671882  0.2236068 -0.1491585  0.19470298  0.000000
#>  [7,] -1.2747549  0.58671882  0.2236068 -0.1491585  0.19470298  0.000000
#>  [8,] -1.2747549  0.02666904  1.3975425  0.8589257 -1.59676515 -1.449138
#>  [9,] -1.2747549  0.58671882  1.3975425  2.3887540 -1.25954762 -1.449138
#> [10,]  0.7844645 -1.09343053 -0.9503289 -0.9774273  0.53192051  0.000000
#> [11,]  0.7844645 -0.53338075 -0.9503289 -1.0053779  2.32338864  0.000000
#> [12,]  0.7844645 -1.09343053 -0.9503289 -1.0482354  0.82698585  0.000000
#> [13,] -1.2747549 -1.09343053 -0.9503289 -0.5917093 -0.26897113 -1.449138
#> [14,] -1.2747549 -0.53338075  1.3975425  1.2520972 -2.25012411 -1.449138
#> [15,]  0.7844645 -1.09343053 -0.9503289 -0.9746322  0.53192051  0.000000
#> [16,]  0.7844645 -0.53338075 -0.9503289 -0.5898459  1.26958386  1.449138
#> [17,]  0.7844645 -0.53338075 -0.9503289 -0.8246308 -0.12143846  1.449138
#> [18,]  0.7844645  0.58671882  1.3975425  1.5595535  0.82698585  1.449138
#> [19,]  0.7844645  1.70681839  0.2236068 -0.3597195 -0.43757989  1.449138
#> [20,]  0.7844645  2.82691796  1.3975425  1.0937106 -0.60618866  1.449138
#> [21,]  0.7844645 -0.53338075 -0.9503289 -0.5833241  0.59514880  0.000000
#>               hp       qsec         vs          wt
#>  [1,] -0.3791310 -0.6870155 -1.0488088 -0.27207115
#>  [2,] -0.3791310 -0.3970190 -1.0488088  0.02031115
#>  [3,] -0.6116907  0.4263639  0.9534626 -0.61605034
#>  [4,]  1.4676668 -1.0080831 -1.0488088  0.81719626
#>  [5,] -0.5843307  2.6479443  0.9534626  0.33562540
#>  [6,] -0.2012912  0.2658301  0.9534626  0.66813861
#>  [7,] -0.2012912  0.5765407  0.9534626  0.66813861
#>  [8,]  0.5784679 -0.0966655 -1.0488088  1.00065182
#>  [9,]  1.2624670 -0.1898787 -1.0488088  2.85240642
#> [10,] -0.9810502  0.8717157  0.9534626 -0.75364201
#> [11,] -1.1725700  0.3797573  0.9534626 -1.42440142
#> [12,] -0.9947302  1.0943916  0.9534626 -1.17215001
#> [13,] -0.5569708  1.1513552  0.9534626 -0.44979373
#> [14,]  0.1680684 -0.4746967 -1.0488088  0.75986639
#> [15,] -0.9810502  0.5765407  0.9534626 -1.05749029
#> [16,] -0.6390507 -0.5627313 -1.0488088 -0.82243785
#> [17,] -0.3380910 -0.4591611  0.9534626 -1.54135434
#> [18,]  1.7275865 -1.7020033 -1.0488088  0.35855735
#> [19,]  0.5100680 -1.1841524 -1.0488088 -0.10008156
#> [20,]  2.6988653 -1.6502182 -1.0488088  0.81719626
#> [21,] -0.3928110  0.4211854  0.9534626 -0.08861559
#> attr(,"center")
#>  [1]   0.6190476   2.9523810   5.6190476 183.6095238   3.8276190   4.0000000
#>  [7] 137.7142857  17.7866667   0.5238095   2.8572857
#> attr(,"scale")
#>  [1]   0.4856209   1.7855555   1.7036708 107.3323051   0.4744712   0.6900656
#>  [7]  73.0995035   1.9310578   0.4994328   0.8721458
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 30
#> 
#> $lambda.min
#> [1] 0.6574149
#> 
#> $null.dev
#> [1] 34.94503
#> 
#> $Bias
#> [1] 10.79856
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 3.403808