Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 40.37878 37.43453 32.99578 29.18983 26.06123 23.50953 21.44816 19.80261
#>   [9] 18.50879 17.51162 16.76384 16.22490 15.86010 15.63974 15.53848 15.51871
#>  [17] 15.50328 15.19799 14.98182 14.78313 14.58699 14.30857 14.07386 13.89071
#>  [25] 13.79081 13.77601 13.80446 13.77832 13.46170 13.16950 12.88241 12.59311
#>  [33] 12.27668 12.06531 11.90695 11.77800 11.55531 11.44979 12.85694 13.84899
#>  [41] 13.92049 14.12141 14.23092 14.43831 15.57885 14.78277 15.22226 16.29051
#>  [49] 16.85632 17.09926 18.92942 18.93271 19.63112 19.63201 19.52279 19.40753
#>  [57] 19.34230 18.72929 18.62175 18.51355 18.62084 18.96470 19.50336 19.50299
#>  [65] 20.05916 20.05836 20.46591 20.46546 20.46500 20.46453 20.46406 20.42932
#>  [73] 20.42891 21.10952 21.30827 21.30857 21.30883 21.30907 21.30928 21.30408
#>  [81] 21.30416 21.08096 21.08088 21.16669 21.16681 21.17551 21.19364 21.19318
#>  [89] 20.85543 20.85154 20.85129 20.85099 20.84295 20.84271 20.84249 20.84630
#>  [97] 20.84610 20.84588 20.84563 20.84354
#> 
#> $cvse
#>   [1] 12.129322 11.643453 10.459128  9.423608  8.569078  7.842877  7.206682
#>   [8]  6.634627  6.111017  5.628111  5.184250  4.782378  4.428826  4.132118
#>  [15]  3.901483  3.745067  3.665923  3.589863  3.586691  3.572654  3.529817
#>  [22]  3.478823  3.436315  3.401835  3.357906  3.325452  3.316712  3.274289
#>  [29]  3.200065  3.155146  3.115496  3.069321  3.027697  3.015377  2.976491
#>  [36]  2.953548  2.954470  2.929190  2.924028  3.362433  3.406380  3.515239
#>  [43]  3.542941  3.595986  4.367670  4.287487  4.226392  4.459536  4.435064
#>  [50]  4.400642  4.570316  4.560917  5.291349  5.291812  5.228438  5.241607
#>  [57]  5.333117  5.358206  5.369706  5.354575  5.349620  5.595465  5.673812
#>  [64]  5.673909  5.829085  5.828963  5.955432  5.955327  5.955207  5.955075
#>  [71]  5.954934  5.952105  5.951956  6.373032  6.353162  6.353337  6.353499
#>  [78]  6.353651  6.353793  6.352290  6.352409  6.344374  6.344476  6.338966
#>  [85]  6.339050  6.337744  6.336591  6.336140  6.063130  6.063393  6.063146
#>  [92]  6.062901  6.064409  6.064186  6.063964  6.063070  6.062857  6.062652
#>  [99]  6.062457  6.061367
#> 
#> $fold
#>  [1]  3  1 10  6  4  8  5  4  1  5  3  6  1  9  9  2  8  2  7 10  7
#> 
#> $lambda
#>   [1] 5.230472308 4.877955980 4.549198073 4.242597348 3.956660487 3.689994812
#>   [7] 3.441301512 3.209369308 2.993068558 2.791345754 2.603218393 2.427770186
#>  [13] 2.264146601 2.111550698 1.969239248 1.836519113 1.712743872 1.597310667
#>  [19] 1.489657274 1.389259360 1.295627928 1.208306942 1.126871098 1.050923758
#>  [25] 0.980095014 0.914039890 0.852436661 0.794985283 0.741405936 0.691437657
#>  [31] 0.644837073 0.601377212 0.560846401 0.523047231 0.487795598 0.454919807
#>  [37] 0.424259735 0.395666049 0.368999481 0.344130151 0.320936931 0.299306856
#>  [43] 0.279134575 0.260321839 0.242777019 0.226414660 0.211155070 0.196923925
#>  [49] 0.183651911 0.171274387 0.159731067 0.148965727 0.138925936 0.129562793
#>  [55] 0.120830694 0.112687110 0.105092376 0.098009501 0.091403989 0.085243666
#>  [61] 0.079498528 0.074140594 0.069143766 0.064483707 0.060137721 0.056084640
#>  [67] 0.052304723 0.048779560 0.045491981 0.042425973 0.039566605 0.036899948
#>  [73] 0.034413015 0.032093693 0.029930686 0.027913458 0.026032184 0.024277702
#>  [79] 0.022641466 0.021115507 0.019692392 0.018365191 0.017127439 0.015973107
#>  [85] 0.014896573 0.013892594 0.012956279 0.012083069 0.011268711 0.010509238
#>  [91] 0.009800950 0.009140399 0.008524367 0.007949853 0.007414059 0.006914377
#>  [97] 0.006448371 0.006013772 0.005608464 0.005230472
#> 
#> $fit
#> $beta
#>               5.2305       4.8780       4.5492      4.2426      3.9567
#> (Intercept) 20.70476 21.599855618 22.532065239 23.40144705 24.21223544
#> am           0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> carb         0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> cyl          0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> disp         0.00000 -0.004543623 -0.009275651 -0.01368876 -0.01780443
#> drat         0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> gear         0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> hp           0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> qsec         0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> vs           0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> wt           0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#>                  3.6900      3.4413      3.2094      2.9931      2.7913
#> (Intercept) 24.96837940 25.67356180 26.33121726 26.94454894 27.51654412
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.02164273 -0.02522233 -0.02856069 -0.03167405 -0.03457757
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  2.6032      2.4278     2.2641      2.1116      1.9692
#> (Intercept) 28.04998874 28.54748097 29.0114439 29.44413728 29.84766857
#> am           0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> disp        -0.03728542 -0.03981076 -0.0421659 -0.04436231 -0.04641069
#> drat         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#>                  1.8365      1.7127      1.5973      1.4897      1.3893
#> (Intercept) 30.22400320 30.52261780 30.52261780 30.52261780 30.52261780
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.04832102 -0.04983683 -0.04983683 -0.04983683 -0.04983683
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  1.2956      1.2083      1.1269      1.0509      0.9801
#> (Intercept) 30.52261780 30.52261780 30.52261780 30.52261780 30.52261780
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.04983683 -0.04983683 -0.04983683 -0.04983683 -0.04983683
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.9140      0.8524      0.7950      0.7414      0.6914
#> (Intercept) 30.52261780 30.31082123 30.01662323 29.74270021 29.48692159
#> am           0.00000000  0.33392296  0.98900236  1.59892774  2.16845828
#> carb         0.00000000 -0.05938156 -0.27372851 -0.47329764 -0.65965166
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.04983683 -0.04867895 -0.04555667 -0.04264963 -0.03993511
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.6448      0.6014      0.5608      0.5230      0.4878
#> (Intercept) 29.24838016 29.02591561 28.89195947 28.89184048 28.89178298
#> am           2.69960760  3.19495924  3.40554925  3.40580876  3.40593418
#> carb        -0.83344712 -0.99552934 -1.03266744 -1.03275039 -1.03279048
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.03740351 -0.03504254 -0.03431503 -0.03431381 -0.03431322
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.4549      0.4243      0.3957      0.3690      0.3441
#> (Intercept) 28.07805107 26.31895805 24.67582673 23.14342539 21.71776058
#> am           3.30766381  3.09547414  2.89734782  2.71257348  2.54058286
#> carb        -1.03305942 -1.03398249 -1.03490208 -1.03575995 -1.03648699
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.03388501 -0.03295129 -0.03207791 -0.03126337 -0.03050711
#> drat         0.20719061  0.65490230  1.07307159  1.46306048  1.82592148
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.3209      0.2993      0.2791      0.2603      0.2428
#> (Intercept) 21.25025761 21.25052833 21.25081934  8.00659925  8.00228956
#> am           2.48391448  2.48425727  2.48446498  3.57277058  3.57346525
#> carb        -1.03683985 -1.03694036 -1.03699081 -0.76664688 -0.76659768
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.03025509 -0.03025377 -0.03025324 -0.02451057 -0.02450734
#> drat         1.94482304  1.94471702  1.94462535  2.12660106  2.12659431
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.55896678  0.55914496
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.2264      0.2112      0.1969      0.1837      0.1713
#> (Intercept)  7.99908648  2.98163084  2.97956231  2.97850260  2.97810456
#> am           3.57398124  3.46629371  3.46704224  3.46758458  3.46797021
#> carb        -0.76656110 -0.75972026 -0.75970925 -0.75971004 -0.75971769
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.02450495 -0.02662742 -0.02662363 -0.02662097 -0.02661913
#> drat         2.12658945  2.50158297  2.50142110  2.50127435  2.50114614
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.55927737  0.83381808  0.83389098  0.83392690  0.83393877
#> vs           0.00000000 -1.73188766 -1.73146285 -1.73112147 -1.73085153
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                 0.1597      0.1490      0.1389      0.1296      0.1208
#> (Intercept)  2.9781209  2.97837939 -7.38143225 -7.38229898 -7.38242924
#> am           3.4682385  3.46842026  4.00358906  4.00417716  4.00461476
#> carb        -0.7597289 -0.75974138 -0.87257552 -0.87271492 -0.87282575
#> cyl          0.0000000  0.00000000  0.87893559  0.87916247  0.87931484
#> disp        -0.0266179 -0.02661711 -0.03279694 -0.03279589 -0.03279497
#> drat         2.5010373  2.50094688  3.05325757  3.05316640  3.05306544
#> gear         0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.8339357  0.83392416  1.07307091  1.07304368  1.07300596
#> vs          -1.7306413 -1.73047991 -1.44226100 -1.44167141 -1.44120548
#> wt           0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.1127     0.1051      0.0980      0.0914      0.0852
#> (Intercept)  2.86662490  2.8690463  2.87093208  2.87237938  2.87347018
#> am           3.34291641  3.3426864  3.34252224  3.34241042  3.34233966
#> carb         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> cyl          0.78842996  0.7882249  0.78807829  0.78797822  0.78791468
#> disp         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> drat         1.47594627  1.4761467  1.47631544  1.47645636  1.47657310
#> gear         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> hp          -0.02747098 -0.0274745 -0.02747778 -0.02748079 -0.02748353
#> qsec         1.20187979  1.2017389  1.20162191  1.20152534  1.20144620
#> vs          -0.90631532 -0.9064817 -0.90659151 -0.90665716 -0.90668882
#> wt          -3.62400188 -3.6235963 -3.62325893 -3.62298030 -3.62275200
#>                  0.0795     0.0741      0.0691      0.0645      0.0601
#> (Intercept)  2.87427348  2.8748470  2.87523873  2.87548834  2.87562838
#> am           3.34230064  3.3422857  3.34228873  3.34230465  3.34232951
#> carb         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> cyl          0.78787940  0.7878656  0.78786783  0.78788158  0.78790334
#> disp         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> drat         1.47666898  1.4767470  1.47680989  1.47685999  1.47689941
#> gear         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> hp          -0.02748599 -0.0274882 -0.02749015 -0.02749186 -0.02749336
#> qsec         1.20138185  1.2013300  1.20128852  1.20125578  1.20123023
#> vs          -0.90669486 -0.9066821 -0.90665594 -0.90662078 -0.90658001
#> wt          -3.62256649 -3.6224172 -3.62229817 -3.62220448 -3.62213172
#>                  0.0561      0.0523      0.0488      0.0455      0.0424
#> (Intercept)  2.87568542  2.87568090  2.87563205  2.87555252  2.87545304
#> am           3.34236018  3.34239422  3.34242977  3.34246544  3.34250023
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.78793032  0.78796034  0.78799175  0.78802331  0.78805411
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         1.47692997  1.47695324  1.47697056  1.47698308  1.47699177
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.02749466 -0.02749577 -0.02749673 -0.02749754 -0.02749823
#> qsec         1.20121059  1.20119578  1.20118486  1.20117706  1.20117173
#> vs          -0.90653627 -0.90649155 -0.90644735 -0.90640471 -0.90636438
#> wt          -3.62207616 -3.62203459 -3.62200433 -3.62198310 -3.62196901
#>                  0.0396      0.0369      0.0344      0.0321      0.0299
#> (Intercept)  2.15608610  2.15489107  2.15390155  2.15310119  2.15247080
#> am           3.24411269  3.24422885  3.24432894  3.24441336  3.24448311
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.84114964  0.84141653  0.84162785  0.84179215  0.84191699
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         1.40420602  1.40386040  1.40357378  1.40333747  1.40314392
#> gear         0.18502611  0.18510831  0.18516284  0.18519539  0.18521069
#> hp          -0.02850273 -0.02850172 -0.02850048 -0.02849909 -0.02849762
#> qsec         1.21316976  1.21326581  1.21335032  1.21342392  1.21348740
#> vs          -0.89152825 -0.89131596 -0.89115067 -0.89102513 -0.89093282
#> wt          -3.64320877 -3.64366657 -3.64405670 -3.64438731 -3.64466577
#>                  0.0279      0.0260      0.0243      0.0226      0.0211
#> (Intercept)  2.15198996  2.15163820  2.15139585  2.15124460  2.15116781
#> am           3.24453956  3.24458424  3.24461867  3.24464437  3.24466273
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.84200901  0.84207401  0.84211698  0.84214223  0.84215342
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         1.40298660  1.40285985  1.40275882  1.40267933  1.40261780
#> gear         0.18521264  0.18520450  0.18518889  0.18516800  0.18514356
#> hp          -0.02849613 -0.02849466 -0.02849324 -0.02849188 -0.02849061
#> qsec         1.21354166  1.21358758  1.21362610  1.21365809  1.21368436
#> vs          -0.89086804 -0.89082577 -0.89080168 -0.89079204 -0.89079367
#> wt          -3.64489876 -3.64509226 -3.64525164 -3.64538169 -3.64548668
#>                  0.0197      0.0184      0.0171      0.0160      0.0149
#> (Intercept)  2.15115066  2.15118019  2.15124525  2.15133638  2.15144568
#> am           3.24467500  3.24468232  3.24468568  3.24468592  3.24468376
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.84215366  0.84214554  0.84213123  0.84211250  0.84209081
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         1.40257113  1.40253671  1.40251230  1.40249599  1.40248617
#> gear         0.18511696  0.18508932  0.18506148  0.18503410  0.18500767
#> hp          -0.02848943 -0.02848834 -0.02848735 -0.02848646 -0.02848566
#> qsec         1.21370570  1.21372279  1.21373627  1.21374669  1.21375457
#> vs          -0.89080390 -0.89082049 -0.89084159 -0.89086568 -0.89089155
#> wt          -3.64557034 -3.64563600 -3.64568655 -3.64572452 -3.64575210
#>                  0.0139      0.0130      0.0121      0.0113      0.0105
#> (Intercept)  2.15156665  2.15169401  2.15182356  2.15195205  2.15207696
#> am           3.24467982  3.24467461  3.24466853  3.24466192  3.24465505
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.84206734  0.84204302  0.84201857  0.84199455  0.84197137
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         1.40248149  1.40248080  1.40248317  1.40248780  1.40249405
#> gear         0.18498253  0.18495892  0.18493699  0.18491680  0.18489839
#> hp          -0.02848494 -0.02848431 -0.02848375 -0.02848327 -0.02848285
#> qsec         1.21376032  1.21376434  1.21376694  1.21376841  1.21376899
#> vs          -0.89091822 -0.89094493 -0.89097110 -0.89099629 -0.89102020
#> wt          -3.64577118 -3.64578337 -3.64579006 -3.64579242 -3.64579143
#>                  0.0098      0.0091     0.0085      0.0079      0.0074
#> (Intercept)  2.15219649  2.15230933  2.1524147  2.15251199  2.15260113
#> am           3.24464813  3.24464132  3.2446347  3.24462849  3.24462262
#> carb         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> cyl          0.84194934  0.84192866  0.8419094  0.84189177  0.84187564
#> disp         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> drat         1.40250140  1.40250943  1.4025178  1.40252624  1.40253456
#> gear         0.18488172  0.18486674  0.1848534  0.18484150  0.18483105
#> hp          -0.02848248 -0.02848217 -0.0284819 -0.02848168 -0.02848149
#> qsec         1.21376887  1.21376823  1.2137672  1.21376589  1.21376441
#> vs          -0.89104262 -0.89106342 -0.8910825 -0.89109996 -0.89111572
#> wt          -3.64578790 -3.64578253 -3.6457759 -3.64576836 -3.64576037
#>                    0.0069       0.0064       0.0060       0.0056       0.0052
#> (Intercept)  2.151826e+00  2.818183637  2.820691047  2.823109273  2.825441496
#> am           3.244617e+00  3.227576753  3.227497306  3.227420698  3.227346826
#> carb         0.000000e+00  0.076427134  0.076585597  0.076738460  0.076885920
#> cyl          8.418610e-01  0.811067789  0.810921809  0.810781002  0.810645188
#> disp         1.658254e-06  0.002585256  0.002590702  0.002595955  0.002601021
#> drat         1.402742e+00  1.305259720  1.304976100  1.304702562  1.304438744
#> gear         1.848084e-01  0.163334128  0.163241741  0.163152603  0.163066607
#> hp          -2.848271e-02 -0.030446548 -0.030449251 -0.030451859 -0.030454374
#> qsec         1.213767e+00  1.226117077  1.226133480  1.226149316  1.226164603
#> vs          -8.911258e-01 -0.848960886 -0.848936367 -0.848912734 -0.848889952
#> wt          -3.645766e+00 -3.881898627 -3.882436038 -3.882954405 -3.883454402
#> 
#> $iter
#>   [1]   0   3   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1   1   1   1  16  19  18  18  18  18   8   1   1  16
#>  [37]  18  18  18  17   9   1   1  31   1   1  35   1   1   1   1   1  42   1
#>  [55]   1 169   1   1   1   1   1   1   1   1   1   1   1   1   1   1  46   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1  93   1   1   1
#> 
#> $lambda
#>   [1] 5.230472308 4.877955980 4.549198073 4.242597348 3.956660487 3.689994812
#>   [7] 3.441301512 3.209369308 2.993068558 2.791345754 2.603218393 2.427770186
#>  [13] 2.264146601 2.111550698 1.969239248 1.836519113 1.712743872 1.597310667
#>  [19] 1.489657274 1.389259360 1.295627928 1.208306942 1.126871098 1.050923758
#>  [25] 0.980095014 0.914039890 0.852436661 0.794985283 0.741405936 0.691437657
#>  [31] 0.644837073 0.601377212 0.560846401 0.523047231 0.487795598 0.454919807
#>  [37] 0.424259735 0.395666049 0.368999481 0.344130151 0.320936931 0.299306856
#>  [43] 0.279134575 0.260321839 0.242777019 0.226414660 0.211155070 0.196923925
#>  [49] 0.183651911 0.171274387 0.159731067 0.148965727 0.138925936 0.129562793
#>  [55] 0.120830694 0.112687110 0.105092376 0.098009501 0.091403989 0.085243666
#>  [61] 0.079498528 0.074140594 0.069143766 0.064483707 0.060137721 0.056084640
#>  [67] 0.052304723 0.048779560 0.045491981 0.042425973 0.039566605 0.036899948
#>  [73] 0.034413015 0.032093693 0.029930686 0.027913458 0.026032184 0.024277702
#>  [79] 0.022641466 0.021115507 0.019692392 0.018365191 0.017127439 0.015973107
#>  [85] 0.014896573 0.013892594 0.012956279 0.012083069 0.011268711 0.010509238
#>  [91] 0.009800950 0.009140399 0.008524367 0.007949853 0.007414059 0.006914377
#>  [97] 0.006448371 0.006013772 0.005608464 0.005230472
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 26
#> 
#> $loss
#>   [1] 753.10952 654.53164 561.87653 484.67043 420.67326 367.95204 324.83992
#>   [8] 289.90048 261.89689 239.76499 222.59009 209.58674 200.08121 193.49631
#>  [15] 189.33823 187.18512 186.66110 186.66110 186.66110 186.66110 186.66110
#>  [22] 186.66110 186.66110 186.66110 186.66110 186.66110 177.04192 155.73440
#>  [29] 140.55987 130.45096 124.55743 122.13492 121.95589 121.95589 121.95589
#>  [36] 119.95032 116.38743 114.01161 112.62453 112.05217 112.01522 112.01522
#>  [43] 112.01522 105.44338 105.44337 105.44337 101.37887 101.37887 101.37886
#>  [50] 101.37886 101.37886 101.37886  96.84870  96.84869  96.84869  85.63213
#>  [57]  85.63212  85.63212  85.63212  85.63211  85.63211  85.63211  85.63211
#>  [64]  85.63211  85.63211  85.63211  85.63211  85.63211  85.63211  85.63211
#>  [71]  85.51543  85.51542  85.51541  85.51541  85.51540  85.51540  85.51540
#>  [78]  85.51540  85.51540  85.51540  85.51540  85.51540  85.51540  85.51540
#>  [85]  85.51540  85.51540  85.51540  85.51540  85.51540  85.51540  85.51540
#>  [92]  85.51540  85.51540  85.51540  85.51540  85.51534  85.45781  85.45779
#>  [99]  85.45778  85.45776
#> 
#> $linear.predictors
#>      5.2305   4.8780   4.5492   4.2426   3.9567   3.6900   3.4413   3.2094
#> 1  20.70476 20.87288 21.04796 21.21125 21.36353 21.50554 21.63799 21.76151
#> 2  20.70476 21.10914 21.53029 21.92306 22.28936 22.63096 22.94955 23.24666
#> 3  20.70476 20.57754 20.44504 20.32148 20.20624 20.09877 19.99854 19.90506
#> 4  20.70476 20.93331 21.17133 21.39331 21.60032 21.79339 21.97345 22.14136
#> 5  20.70476 20.96011 21.22605 21.47407 21.70537 21.92108 22.12226 22.30987
#> 6  20.70476 20.83834 20.97747 21.10721 21.22821 21.34106 21.44630 21.54445
#> 7  20.70476 20.83834 20.97747 21.10721 21.22821 21.34106 21.44630 21.54445
#> 8  20.70476 20.34672 19.97384 19.62609 19.30177 18.99931 18.71724 18.45418
#> 9  20.70476 20.34672 19.97384 19.62609 19.30177 18.99931 18.71724 18.45418
#> 10 20.70476 20.34672 19.97384 19.62609 19.30177 18.99931 18.71724 18.45418
#> 11 20.70476 19.50979 18.26527 17.10462 16.02220 15.01272 14.07129 13.19330
#> 12 20.70476 21.24227 21.80207 22.32414 22.81103 23.26510 23.68856 24.08349
#> 13 20.70476 21.25590 21.82990 22.36521 22.86444 23.33002 23.76423 24.16917
#> 14 20.70476 21.27680 21.87257 22.42818 22.94634 23.42958 23.88025 24.30055
#> 15 20.70476 21.05417 21.41806 21.75743 22.07392 22.36909 22.64436 22.90108
#> 16 20.70476 20.00959 19.28559 18.61038 17.98068 17.39342 16.84574 16.33498
#> 17 20.70476 21.05326 21.41620 21.75469 22.07036 22.36476 22.63931 22.89537
#> 18 20.70476 20.00504 19.27631 18.59669 17.96288 17.37178 16.82052 16.30642
#> 19 20.70476 20.94103 21.18710 21.41658 21.63059 21.83018 22.01632 22.18992
#> 20 20.70476 20.23223 19.74009 19.28113 18.85310 18.45392 18.08164 17.73445
#> 21 20.70476 21.05008 21.40971 21.74511 22.05790 22.34961 22.62166 22.87537
#>      2.9931   2.7913   2.6032   2.4278    2.2641    2.1116    1.9692    1.8365
#> 1  21.87670 21.98413 22.08432 22.17776 22.264900 22.346167 22.421958 22.492640
#> 2  23.52375 23.78217 24.02316 24.24792 24.457527 24.653008 24.835314 25.005333
#> 3  19.81789 19.73659 19.66077 19.59006 19.524117 19.462617 19.405262 19.351773
#> 4  22.29797 22.44401 22.58022 22.70724 22.825707 22.936186 23.039220 23.135309
#> 5  22.48484 22.64802 22.80020 22.94213 23.074485 23.197924 23.313043 23.420403
#> 6  21.63598 21.72134 21.80095 21.87520 21.944439 22.009014 22.069236 22.125400
#> 7  21.63598 21.72134 21.80095 21.87520 21.944439 22.009014 22.069236 22.125400
#> 8  18.20885 17.98005 17.76667 17.56767 17.382089 17.209012 17.047599 16.897065
#> 9  18.20885 17.98005 17.76667 17.56767 17.382089 17.209012 17.047599 16.897065
#> 10 18.20885 17.98005 17.76667 17.56767 17.382089 17.209012 17.047599 16.897065
#> 11 12.37449 11.61086 10.89870 10.23453  9.615131  9.037474  8.498749  7.996333
#> 12 24.45180 24.79529 25.11563 25.41437 25.692988 25.952823 26.195147 26.421139
#> 13 24.54682 24.89902 25.22748 25.53381 25.819485 26.085910 26.334379 26.566102
#> 14 24.69252 25.05808 25.39900 25.71694 26.013449 26.289977 26.547868 26.788379
#> 15 23.14050 23.36378 23.57201 23.76621 23.947319 24.116224 24.273744 24.420648
#> 16 15.85863 15.41439 15.00009 14.61372 14.253379 13.917328 13.603926 13.311646
#> 17 23.13416 23.35686 23.56455 23.75825 23.938886 24.107351 24.264462 24.410984
#> 18 15.82696 15.37982 14.96281 14.57391 14.211214 13.872966 13.557515 13.263325
#> 19 22.35181 22.50280 22.64360 22.77492 22.897389 23.011602 23.118118 23.217455
#> 20 17.41066 17.10869 16.82708 16.56444 16.319508 16.091082 15.878050 15.679376
#> 21 23.11199 23.33266 23.53845 23.73038 23.909370 24.076298 24.231975 24.377160
#>       1.7127    1.5973    1.4897    1.3893    1.2956    1.2083    1.1269
#> 1  22.548725 22.548725 22.548725 22.548725 22.548725 22.548725 22.548725
#> 2  25.140240 25.140240 25.140240 25.140240 25.140240 25.140240 25.140240
#> 3  19.309331 19.309331 19.309331 19.309331 19.309331 19.309331 19.309331
#> 4  23.211555 23.211555 23.211555 23.211555 23.211555 23.211555 23.211555
#> 5  23.505592 23.505592 23.505592 23.505592 23.505592 23.505592 23.505592
#> 6  22.169965 22.169965 22.169965 22.169965 22.169965 22.169965 22.169965
#> 7  22.169965 22.169965 22.169965 22.169965 22.169965 22.169965 22.169965
#> 8  16.777620 16.777620 16.777620 16.777620 16.777620 16.777620 16.777620
#> 9  16.777620 16.777620 16.777620 16.777620 16.777620 16.777620 16.777620
#> 10 16.777620 16.777620 16.777620 16.777620 16.777620 16.777620 16.777620
#> 11  7.597675  7.597675  7.597675  7.597675  7.597675  7.597675  7.597675
#> 12 26.600459 26.600459 26.600459 26.600459 26.600459 26.600459 26.600459
#> 13 26.749970 26.749970 26.749970 26.749970 26.749970 26.749970 26.749970
#> 14 26.979219 26.979219 26.979219 26.979219 26.979219 26.979219 26.979219
#> 15 24.537214 24.537214 24.537214 24.537214 24.537214 24.537214 24.537214
#> 16 13.079727 13.079727 13.079727 13.079727 13.079727 13.079727 13.079727
#> 17 24.527247 24.527247 24.527247 24.527247 24.527247 24.527247 24.527247
#> 18 13.029890 13.029890 13.029890 13.029890 13.029890 13.029890 13.029890
#> 19 23.296277 23.296277 23.296277 23.296277 23.296277 23.296277 23.296277
#> 20 15.521731 15.521731 15.521731 15.521731 15.521731 15.521731 15.521731
#> 21 24.492361 24.492361 24.492361 24.492361 24.492361 24.492361 24.492361
#>       1.0509    0.9801    0.9140    0.8524    0.7950    0.7414    0.6914
#> 1  22.548725 22.548725 22.548725 22.618586 22.621644 22.624496 22.627156
#> 2  25.140240 25.140240 25.140240 25.328036 25.811777 26.262170 26.682737
#> 3  19.309331 19.309331 19.309331 19.298676 19.492644 19.673235 19.841871
#> 4  23.211555 23.211555 23.211555 23.050856 22.786003 22.539404 22.309138
#> 5  23.505592 23.505592 23.505592 23.338062 23.054787 22.791037 22.544755
#> 6  22.169965 22.169965 22.169965 21.914703 21.286411 20.701431 20.155191
#> 7  22.169965 22.169965 22.169965 21.914703 21.286411 20.701431 20.155191
#> 8  16.777620 16.777620 16.777620 16.707022 16.630908 16.560039 16.493864
#> 9  16.777620 16.777620 16.777620 16.707022 16.630908 16.560039 16.493864
#> 10 16.777620 16.777620 16.777620 16.707022 16.630908 16.560039 16.493864
#> 11  7.597675  7.597675  7.597675  7.680978  7.965641  8.230679  8.478166
#> 12 26.600459 26.600459 26.600459 26.754329 27.146587 27.511804 27.852835
#> 13 26.749970 26.749970 26.749970 26.840985 27.009529 27.166456 27.312989
#> 14 26.979219 26.979219 26.979219 27.124289 27.492818 27.835941 28.156342
#> 15 24.537214 24.537214 24.537214 24.405098 24.271539 24.147182 24.031064
#> 16 13.079727 13.079727 13.079727 13.035662 12.976875 12.922138 12.871028
#> 17 24.527247 24.527247 24.527247 24.669903 24.977701 25.264282 25.531883
#> 18 13.029890 13.029890 13.029890 13.320906 13.920320 14.478417 14.999551
#> 19 23.296277 23.296277 23.296277 23.230007 22.757537 22.317645 21.906880
#> 20 15.521731 15.521731 15.521731 15.517328 15.103240 14.717708 14.357700
#> 21 24.492361 24.492361 24.492361 24.635828 24.945811 25.234427 25.503929
#>       0.6448    0.6014    0.5608    0.5230    0.4878    0.4549   0.4243
#> 1  22.629637 22.631951 22.676435 22.676439 22.676441 22.639919 22.56041
#> 2  27.074961 27.440751 27.558819 27.559008 27.559099 27.490758 27.34308
#> 3  19.999142 20.145814 20.138411 20.138484 20.138519 19.992710 19.67846
#> 4  22.094390 21.894116 21.792610 21.792504 21.792453 21.805535 21.83363
#> 5  22.315071 22.100867 21.995069 21.994956 21.994901 22.053110 22.17867
#> 6  19.645763 19.170668 19.010091 19.009845 19.009726 19.078873 19.22761
#> 7  19.645763 19.170668 19.010091 19.009845 19.009726 19.078873 19.22761
#> 8  16.432150 16.374594 16.329873 16.329842 16.329827 16.269462 16.13959
#> 9  16.432150 16.374594 16.329873 16.329842 16.329827 16.269462 16.13959
#> 10 16.432150 16.374594 16.329873 16.329842 16.329827 16.269462 16.13959
#> 11  8.708975  8.924228  8.976378  8.976488  8.976542  8.980281  8.99014
#> 12 28.170884 28.467497 28.564249 28.564402 28.564477 28.531243 28.45918
#> 13 27.449648 27.577096 27.634526 27.634593 27.634626 27.775950 28.08072
#> 14 28.455151 28.733821 28.825043 28.825187 28.825257 28.817776 28.80130
#> 15 23.922771 23.821777 23.738057 23.738002 23.737975 23.742007 23.75066
#> 16 12.823362 12.778908 12.751031 12.751007 12.750995 12.858881 13.09286
#> 17 25.781451 26.014198 26.104076 26.104198 26.104256 26.161084 26.28364
#> 18 15.485566 15.938825 16.122265 16.122502 16.122616 16.234183 16.47629
#> 19 21.523796 21.166530 21.125825 21.125645 21.125558 21.024062 20.80335
#> 20 14.021953 13.708835 13.707347 13.707191 13.707115 13.655306 13.54259
#> 21 25.755268 25.989668 26.080056 26.080178 26.080237 26.071063 26.05101
#>       0.3957    0.3690    0.3441    0.3209    0.2993   0.2791   0.2603   0.2428
#> 1  22.486080 22.416756 22.352351 22.330809 22.330818 22.33084 22.39845 22.39855
#> 2  27.205184 27.076578 26.956886 26.917351 26.917599 26.91775 26.75537 26.75544
#> 3  19.385073 19.111454 18.856717 18.773735 18.773910 18.77401 18.89680 18.89685
#> 4  21.859828 21.884262 21.907043 21.914554 21.914426 21.91435 21.90410 21.90390
#> 5  22.295894 22.405220 22.506997 22.540368 22.540208 22.54011 24.15883 24.15913
#> 6  19.366402 19.495842 19.616433 19.655852 19.655526 19.65534 19.39741 19.39707
#> 7  19.366402 19.495842 19.616433 19.655852 19.655526 19.65534 19.73279 19.73256
#> 8  16.018364 15.905304 15.800017 15.765992 15.766000 15.76600 15.20133 15.20114
#> 9  16.018364 15.905304 15.800017 15.765992 15.766000 15.76600 15.31312 15.31297
#> 10 16.018364 15.905304 15.800017 15.765992 15.766000 15.76600 15.53671 15.53662
#> 11  8.999596  9.008417  9.016305  9.020027  9.020185  9.02024 10.00574 10.00627
#> 12 28.391873 28.329098 28.270706 28.251135 28.251319 28.25143 28.44336 28.44349
#> 13 28.365316 28.630730 28.877774 28.958160 28.958150 28.95814 29.02683 29.02683
#> 14 28.785895 28.771529 28.758189 28.753349 28.753508 28.75361 29.16772 29.16790
#> 15 23.758733 23.766258 23.773279 23.775627 23.775564 23.77553 23.34958 23.34925
#> 16 13.311508 13.515421 13.705010 13.767807 13.767743 13.76767 12.90721 12.90695
#> 17 26.398105 26.504853 26.604196 26.636371 26.636473 26.63653 25.85304 25.85286
#> 18 16.702583 16.913631 17.109788 17.174430 17.174658 17.17475 16.98885 16.98912
#> 19 20.596985 20.404529 20.225726 20.166405 20.166223 20.16616 19.78774 19.78762
#> 20 13.437182 13.338879 13.247569 13.217345 13.217177 13.21711 13.75760 13.75792
#> 21 26.032268 26.014790 25.998546 25.992849 25.992986 25.99307 26.21741 26.21757
#>      0.2264   0.2112   0.1969   0.1837   0.1713   0.1597   0.1490    0.1389
#> 1  22.39863 23.09641 23.09635 23.09630 23.09625 23.09621 23.09618 23.329330
#> 2  26.75549 26.22900 26.22926 26.22948 26.22965 26.22979 26.22990 26.005885
#> 3  18.89689 18.26303 18.26327 18.26347 18.26364 18.26377 18.26387 18.322519
#> 4  21.90375 21.73126 21.73106 21.73091 21.73080 21.73072 21.73066 20.863526
#> 5  24.15935 24.88180 24.88175 24.88165 24.88153 24.88142 24.88130 24.871183
#> 6  19.39682 18.81318 18.81292 18.81272 18.81259 18.81249 18.81243 19.068819
#> 7  19.73239 19.31347 19.31325 19.31308 19.31295 19.31286 19.31278 19.712662
#> 8  15.20099 15.54692 15.54670 15.54655 15.54645 15.54638 15.54634 16.031866
#> 9  15.31285 15.71369 15.71348 15.71334 15.71324 15.71317 15.71313 16.246480
#> 10 15.53656 16.04721 16.04704 16.04691 16.04681 16.04674 16.04670 16.675708
#> 11 10.00666 10.05753 10.05805 10.05842 10.05866 10.05882 10.05891  9.355057
#> 12 28.44358 28.30164 28.30181 28.30194 28.30204 28.30212 28.30218 28.591925
#> 13 29.02682 28.95602 28.95598 28.95595 28.95592 28.95589 28.95587 29.393592
#> 14 29.16803 29.21277 29.21292 29.21303 29.21310 29.21315 29.21319 29.730059
#> 15 23.34900 23.23263 23.23231 23.23209 23.23193 23.23183 23.23177 22.649764
#> 16 12.90675 12.80320 12.80301 12.80289 12.80281 12.80276 12.80273 12.605497
#> 17 25.85272 26.73198 26.73164 26.73139 26.73121 26.73109 26.73101 25.893492
#> 18 16.98932 16.70986 16.71029 16.71061 16.71084 16.71101 16.71112 17.095890
#> 19 19.78754 20.00854 20.00838 20.00827 20.00819 20.00815 20.00812 19.590153
#> 20 13.75817 13.38466 13.38506 13.38534 13.38554 13.38567 13.38576 13.276527
#> 21 26.21769 25.76520 25.76548 25.76569 25.76584 25.76596 25.76605 25.490066
#>       0.1296    0.1208    0.1127    0.1051    0.0980    0.0914    0.0852
#> 1  23.329203 23.329093 23.711492 23.711616 23.711734 23.711845 23.711947
#> 2  26.006219 26.006500 25.543836 25.543805 25.543780 25.543762 25.543747
#> 3  18.322897 18.323201 19.643011 19.642774 19.642594 19.642459 19.642361
#> 4  20.863150 20.862882 21.334100 21.334533 21.334878 21.335150 21.335361
#> 5  24.870701 24.870294 24.397437 24.397375 24.397298 24.397212 24.397121
#> 6  19.068665 19.068540 18.625501 18.625696 18.625871 18.626025 18.626161
#> 7  19.712492 19.712343 19.346629 19.346740 19.346844 19.346940 19.347029
#> 8  16.031931 16.031974 14.923463 14.923426 14.923405 14.923395 14.923395
#> 9  16.246539 16.246575 16.396000 16.395796 16.395637 16.395514 16.395420
#> 10 16.675757 16.675777 16.695552 16.695312 16.695123 16.694975 16.694861
#> 11  9.355169  9.355262  9.456554  9.456869  9.457129  9.457341  9.457513
#> 12 28.592185 28.592382 28.093517 28.093457 28.093419 28.093397 28.093388
#> 13 29.393657 29.393692 30.710920 30.710977 30.711041 30.711111 30.711181
#> 14 29.730285 29.730446 30.167190 30.166953 30.166768 30.166626 30.166518
#> 15 22.649498 22.649314 23.026795 23.026811 23.026798 23.026763 23.026713
#> 16 12.605493 12.605503 12.553754 12.553807 12.553839 12.553855 12.553860
#> 17 25.893109 25.892838 25.717872 25.718327 25.718679 25.718948 25.719149
#> 18 17.096457 17.096889 17.432306 17.432017 17.431786 17.431603 17.431461
#> 19 19.589799 19.589539 20.066276 20.066287 20.066288 20.066281 20.066268
#> 20 13.276543 13.276551 13.148810 13.148283 13.147828 13.147437 13.147104
#> 21 25.490251 25.490406 23.808987 23.809139 23.809263 23.809361 23.809440
#>       0.0795   0.0741    0.0691    0.0645    0.0601    0.0561    0.0523
#> 1  23.712041 23.71212 23.712200 23.712267 23.712326 23.712378 23.712422
#> 2  25.543737 25.54373 25.543725 25.543722 25.543721 25.543720 25.543721
#> 3  19.642293 19.64225 19.642222 19.642210 19.642208 19.642214 19.642225
#> 4  21.335523 21.33564 21.335733 21.335795 21.335837 21.335862 21.335875
#> 5  24.397029 24.39694 24.396853 24.396771 24.396696 24.396626 24.396563
#> 6  18.626279 18.62638 18.626471 18.626547 18.626611 18.626666 18.626712
#> 7  19.347108 19.34718 19.347244 19.347300 19.347349 19.347392 19.347429
#> 8  14.923402 14.92341 14.923428 14.923445 14.923462 14.923479 14.923497
#> 9  16.395351 16.39530 16.395267 16.395245 16.395233 16.395227 16.395227
#> 10 16.694776 16.69471 16.694668 16.694637 16.694618 16.694608 16.694604
#> 11  9.457651  9.45776  9.457846  9.457912  9.457962  9.457999  9.458025
#> 12 28.093389 28.09340 28.093409 28.093425 28.093442 28.093460 28.093479
#> 13 30.711250 30.71132 30.711380 30.711438 30.711492 30.711540 30.711584
#> 14 30.166440 30.16638 30.166345 30.166321 30.166308 30.166304 30.166305
#> 15 23.026654 23.02659 23.026525 23.026459 23.026396 23.026336 23.026280
#> 16 12.553855 12.55384 12.553828 12.553809 12.553789 12.553769 12.553749
#> 17 25.719294 25.71940 25.719463 25.719504 25.719523 25.719527 25.719519
#> 18 17.431351 17.43127 17.431208 17.431164 17.431135 17.431116 17.431105
#> 19 20.066253 20.06624 20.066217 20.066198 20.066180 20.066163 20.066146
#> 20 13.146822 13.14658 13.146386 13.146221 13.146085 13.145973 13.145883
#> 21 23.809501 23.80955 23.809583 23.809609 23.809628 23.809641 23.809649
#>       0.0488    0.0455    0.0424    0.0396    0.0369   0.0344   0.0321   0.0299
#> 1  23.712461 23.712494 23.712522 23.702229 23.702161 23.70209 23.70203 23.70196
#> 2  25.543722 25.543724 25.543726 25.493658 25.493676 25.49369 25.49371 25.49373
#> 3  19.642240 19.642257 19.642276 19.674146 19.674521 19.67483 19.67507 19.67527
#> 4  21.335879 21.335876 21.335868 21.425171 21.424832 21.42454 21.42428 21.42406
#> 5  24.396506 24.396456 24.396411 24.471469 24.471381 24.47132 24.47128 24.47126
#> 6  18.626750 18.626782 18.626808 18.718581 18.718480 18.71838 18.71827 18.71817
#> 7  19.347461 19.347488 19.347511 19.446483 19.446439 19.44639 19.44632 19.44626
#> 8  14.923513 14.923529 14.923543 14.902077 14.902192 14.90228 14.90235 14.90241
#> 9  16.395231 16.395238 16.395247 16.383402 16.383691 16.38393 16.38413 16.38429
#> 10 16.694605 16.694610 16.694617 16.686510 16.686814 16.68707 16.68728 16.68745
#> 11  9.458043  9.458055  9.458061  9.382814  9.382408  9.38207  9.38179  9.38156
#> 12 28.093497 28.093514 28.093530 28.066710 28.066759 28.06680 28.06683 28.06685
#> 13 30.711623 30.711657 30.711687 30.638089 30.638007 30.63793 30.63787 30.63781
#> 14 30.166310 30.166319 30.166330 30.143236 30.143444 30.14362 30.14377 30.14389
#> 15 23.026228 23.026181 23.026139 22.910050 22.909993 22.90997 22.90997 22.90998
#> 16 12.553729 12.553711 12.553694 12.399906 12.399772 12.39968 12.39961 12.39957
#> 17 25.719504 25.719483 25.719458 25.780281 25.779865 25.77951 25.77922 25.77897
#> 18 17.431101 17.431102 17.431106 17.497545 17.497761 17.49794 17.49809 17.49821
#> 19 20.066131 20.066117 20.066105 20.179919 20.179997 20.18006 20.18010 20.18014
#> 20 13.145810 13.145752 13.145707 13.183025 13.183374 13.18369 13.18397 13.18423
#> 21 23.809653 23.809655 23.809654 23.714700 23.714433 23.71422 23.71404 23.71391
#>       0.0279    0.0260    0.0243    0.0226    0.0211    0.0197    0.0184
#> 1  23.701902 23.701845 23.701793 23.701745 23.701701 23.701662 23.701627
#> 2  25.493751 25.493768 25.493785 25.493800 25.493813 25.493826 25.493837
#> 3  19.675425 19.675545 19.675636 19.675703 19.675750 19.675782 19.675800
#> 4  21.423875 21.423718 21.423586 21.423478 21.423389 21.423317 21.423259
#> 5  24.471256 24.471259 24.471270 24.471285 24.471305 24.471326 24.471349
#> 6  18.718070 18.717977 18.717890 18.717809 18.717735 18.717668 18.717607
#> 7  19.446195 19.446129 19.446066 19.446004 19.445946 19.445891 19.445841
#> 8  14.902451 14.902483 14.902505 14.902520 14.902530 14.902535 14.902537
#> 9  16.384425 16.384531 16.384616 16.384682 16.384732 16.384770 16.384797
#> 10 16.687597 16.687712 16.687804 16.687876 16.687932 16.687974 16.688005
#> 11  9.381372  9.381221  9.381101  9.381006  9.380934  9.380879  9.380839
#> 12 28.066867 28.066878 28.066885 28.066888 28.066889 28.066887 28.066884
#> 13 30.637753 30.637705 30.637663 30.637625 30.637592 30.637563 30.637538
#> 14 30.143992 30.144075 30.144141 30.144193 30.144233 30.144263 30.144286
#> 15 22.910014 22.910056 22.910105 22.910159 22.910215 22.910271 22.910327
#> 16 12.399553 12.399549 12.399557 12.399574 12.399598 12.399626 12.399656
#> 17 25.778773 25.778611 25.778481 25.778379 25.778301 25.778242 25.778199
#> 18 17.498314 17.498392 17.498453 17.498499 17.498532 17.498555 17.498570
#> 19 20.180161 20.180178 20.180189 20.180195 20.180198 20.180198 20.180197
#> 20 13.184452 13.184648 13.184819 13.184966 13.185093 13.185200 13.185291
#> 21 23.713801 23.713719 23.713658 23.713614 23.713584 23.713564 23.713554
#>       0.0171    0.0160    0.0149    0.0139    0.0130    0.0121    0.0113
#> 1  23.701596 23.701568 23.701544 23.701523 23.701505 23.701490 23.701477
#> 2  25.493848 25.493857 25.493865 25.493872 25.493879 25.493884 25.493889
#> 3  19.675809 19.675810 19.675805 19.675796 19.675783 19.675768 19.675752
#> 4  21.423214 21.423179 21.423153 21.423135 21.423122 21.423113 21.423109
#> 5  24.471372 24.471395 24.471417 24.471438 24.471458 24.471476 24.471493
#> 6  18.717553 18.717504 18.717461 18.717423 18.717390 18.717361 18.717337
#> 7  19.445794 19.445752 19.445714 19.445679 19.445649 19.445622 19.445598
#> 8  14.902535 14.902532 14.902528 14.902523 14.902516 14.902510 14.902503
#> 9  16.384816 16.384828 16.384835 16.384837 16.384836 16.384832 16.384827
#> 10 16.688026 16.688041 16.688049 16.688052 16.688052 16.688049 16.688044
#> 11  9.380812  9.380794  9.380785  9.380781  9.380782  9.380787  9.380795
#> 12 28.066880 28.066875 28.066869 28.066863 28.066857 28.066851 28.066845
#> 13 30.637515 30.637496 30.637479 30.637465 30.637453 30.637443 30.637434
#> 14 30.144301 30.144311 30.144317 30.144319 30.144318 30.144316 30.144312
#> 15 22.910381 22.910432 22.910480 22.910526 22.910567 22.910605 22.910640
#> 16 12.399688 12.399721 12.399753 12.399783 12.399813 12.399840 12.399866
#> 17 25.778170 25.778152 25.778142 25.778139 25.778141 25.778147 25.778156
#> 18 17.498578 17.498581 17.498580 17.498575 17.498568 17.498560 17.498550
#> 19 20.180194 20.180190 20.180185 20.180180 20.180175 20.180170 20.180165
#> 20 13.185367 13.185430 13.185482 13.185524 13.185558 13.185585 13.185606
#> 21 23.713550 23.713552 23.713558 23.713566 23.713577 23.713589 23.713601
#>       0.0105    0.0098    0.0091    0.0085    0.0079    0.0074    0.0069
#> 1  23.701465 23.701456 23.701448 23.701442 23.701436 23.701432 23.701435
#> 2  25.493893 25.493897 25.493900 25.493902 25.493904 25.493906 25.493859
#> 3  19.675736 19.675719 19.675702 19.675686 19.675671 19.675657 19.675560
#> 4  21.423108 21.423108 21.423111 21.423115 21.423120 21.423125 21.423151
#> 5  24.471509 24.471523 24.471536 24.471547 24.471557 24.471566 24.471596
#> 6  18.717315 18.717297 18.717282 18.717269 18.717258 18.717249 18.717248
#> 7  19.445577 19.445558 19.445543 19.445529 19.445517 19.445508 19.445508
#> 8  14.902497 14.902491 14.902485 14.902479 14.902474 14.902469 14.902399
#> 9  16.384820 16.384812 16.384804 16.384796 16.384788 16.384780 16.384713
#> 10 16.688038 16.688030 16.688022 16.688014 16.688006 16.687998 16.687931
#> 11  9.380804  9.380815  9.380826  9.380838  9.380849  9.380861  9.381033
#> 12 28.066840 28.066835 28.066830 28.066826 28.066822 28.066818 28.066805
#> 13 30.637427 30.637421 30.637416 30.637412 30.637409 30.637407 30.637584
#> 14 30.144306 30.144301 30.144294 30.144288 30.144282 30.144275 30.144284
#> 15 22.910671 22.910699 22.910724 22.910745 22.910764 22.910781 22.910749
#> 16 12.399890 12.399911 12.399931 12.399949 12.399964 12.399978 12.400087
#> 17 25.778168 25.778180 25.778193 25.778207 25.778220 25.778233 25.778313
#> 18 17.498540 17.498530 17.498519 17.498509 17.498500 17.498491 17.498630
#> 19 20.180160 20.180155 20.180151 20.180147 20.180143 20.180140 20.179955
#> 20 13.185623 13.185634 13.185643 13.185649 13.185652 13.185654 13.185480
#> 21 23.713614 23.713627 23.713640 23.713652 23.713663 23.713674 23.713680
#>       0.0064    0.0060    0.0056    0.0052
#> 1  23.734300 23.734318 23.734336 23.734353
#> 2  25.455797 25.455758 25.455720 25.455683
#> 3  19.750089 19.750257 19.750419 19.750576
#> 4  21.466749 21.466843 21.466933 21.467021
#> 5  24.457985 24.457962 24.457939 24.457917
#> 6  18.683867 18.683708 18.683554 18.683405
#> 7  19.419538 19.419388 19.419243 19.419104
#> 8  14.800902 14.800682 14.800471 14.800267
#> 9  16.365971 16.365937 16.365905 16.365874
#> 10 16.662323 16.662269 16.662217 16.662167
#> 11  9.455414  9.455561  9.455702  9.455839
#> 12 28.022604 28.022492 28.022383 28.022278
#> 13 30.733097 30.733222 30.733343 30.733460
#> 14 30.160262 30.160275 30.160287 30.160298
#> 15 22.932281 22.932456 22.932625 22.932788
#> 16 12.404465 12.404536 12.404604 12.404670
#> 17 25.751120 25.751095 25.751072 25.751049
#> 18 17.507493 17.507460 17.507429 17.507398
#> 19 20.211113 20.211209 20.211302 20.211392
#> 20 13.204510 13.204626 13.204738 13.204846
#> 21 23.620121 23.619946 23.619778 23.619616
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 22.8 18.1 24.4 22.8 19.2 17.8 16.4 17.3 15.2 10.4 32.4 30.4 33.9 21.5
#> [16] 13.3 26.0 15.8 19.7 15.0 21.4
#> 
#> $X
#>               am        carb        cyl       disp        drat       gear
#>  [1,]  1.0488088  0.59568340  0.1097643 -0.3550435  0.28914467  0.2013468
#>  [2,]  1.0488088 -1.11013724 -1.0427605 -0.8540236  0.19276312  0.2013468
#>  [3,] -0.9534626 -1.11013724  0.1097643  0.2686816 -1.90835485 -1.2080809
#>  [4,] -0.9534626 -0.54153036 -1.0427605 -0.4826673 -0.11565787  0.2013468
#>  [5,] -0.9534626 -0.54153036 -1.0427605 -0.5392823  0.32769730  0.2013468
#>  [6,] -0.9534626  0.59568340  0.1097643 -0.2821157  0.32769730  0.2013468
#>  [7,] -0.9534626  0.59568340  0.1097643 -0.2821157  0.32769730  0.2013468
#>  [8,] -0.9534626  0.02707652  1.2622890  0.7561468 -1.31078919 -1.2080809
#>  [9,] -0.9534626  0.02707652  1.2622890  0.7561468 -1.31078919 -1.2080809
#> [10,] -0.9534626  0.02707652  1.2622890  0.7561468 -1.31078919 -1.2080809
#> [11,] -0.9534626  0.59568340  1.2622890  2.5236879 -1.44572337 -1.2080809
#> [12,]  1.0488088 -1.11013724 -1.0427605 -1.1351797  0.63611828  0.2013468
#> [13,]  1.0488088 -0.54153036 -1.0427605 -1.1639671  2.27460477  0.2013468
#> [14,]  1.0488088 -1.11013724 -1.0427605 -1.2081076  0.90598665  0.2013468
#> [15,] -0.9534626 -1.11013724 -1.0427605 -0.7379148 -0.09638156 -1.2080809
#> [16,] -0.9534626  0.59568340  1.2622890  1.4681530 -0.03855262 -1.2080809
#> [17,]  1.0488088 -0.54153036 -1.0427605 -0.7359957  1.31078919  1.6107745
#> [18,]  1.0488088  0.59568340  1.2622890  1.4777488  0.90598665  1.6107745
#> [19,]  1.0488088  1.73289716  0.1097643 -0.4989801 -0.25059205  1.6107745
#> [20,]  1.0488088  2.87011091  1.2622890  0.9979602 -0.40480254  1.6107745
#> [21,]  1.0488088 -0.54153036 -1.0427605 -0.7292786  0.69394722  0.2013468
#>               hp        qsec         vs          wt
#>  [1,] -0.4249675 -0.53575388 -1.0488088 -0.27404898
#>  [2,] -0.6565910  0.25719036  0.9534626 -0.92736067
#>  [3,] -0.4930921  1.06010875  0.9534626  0.41457686
#>  [4,] -1.0789633  0.95039319  0.9534626  0.09674955
#>  [5,] -0.6293412  2.39664371  0.9534626  0.04966402
#>  [6,] -0.2478437  0.10259117  0.9534626  0.39103409
#>  [7,] -0.2478437  0.40181541  0.9534626  0.39103409
#>  [8,]  0.5287764 -0.34624520 -1.0488088  1.13263115
#>  [9,]  0.5287764 -0.24650378 -1.0488088  0.73240417
#> [10,]  0.5287764 -0.04702095 -1.0488088  0.79126107
#> [11,]  1.0056483 -0.13678823 -1.0488088  2.72647625
#> [12,] -1.0244637  0.68607844  0.9534626 -1.06861725
#> [13,] -1.2152125  0.21230673  0.9534626 -1.75724309
#> [14,] -1.0380886  0.90052249  0.9534626 -1.49827269
#> [15,] -0.6020914  0.95538026  0.9534626 -0.75667563
#> [16,]  1.4143957 -1.33867227 -1.0488088  0.86188937
#> [17,] -0.6838409 -0.69534015 -1.0488088 -1.13924554
#> [18,]  1.6732691 -1.79249571 -1.0488088  0.07320678
#> [19,]  0.4606518 -1.29378864 -1.0488088 -0.39764849
#> [20,]  2.6406379 -1.74262500 -1.0488088  0.54406206
#> [21,] -0.4385924  0.25220329  0.9534626 -0.38587711
#> attr(,"center")
#>  [1]   0.4761905   2.9523810   5.8095238 197.0000000   3.7500000   3.8571429
#>  [7] 141.1904762  18.0942857   0.5238095   3.1078095
#> attr(,"scale")
#>  [1]   0.4994328   1.7586843   1.7353207 104.2125712   0.5187714   0.7095078
#>  [7]  73.3949648   2.0051851   0.4994328   0.8495179
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 38
#> 
#> $lambda.min
#> [1] 0.395666
#> 
#> $null.dev
#> [1] 35.86236
#> 
#> $Bias
#> [1] 6.097247
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 10.26066