Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 37.454570 36.246442 31.615983 26.706227 22.622101 19.243520 16.466889
#>   [8] 14.202892 12.374582 10.915722  9.769341  8.886483  8.225129  7.749249
#>  [15]  7.427989  7.216272  6.989609  6.786397  6.725801  6.730472  6.783895
#>  [22]  6.848826  6.918418  6.955154  6.922296  6.880728  6.734757  6.610765
#>  [29]  6.514299  6.441010  6.387209  6.349603  6.325503  6.312595  6.323392
#>  [36]  6.301293  6.258590  6.218090  6.677995  7.170240  7.463422  8.396717
#>  [43]  8.413068  8.459558  8.505233  9.017241  9.423565 12.862184 13.135993
#>  [50] 13.220341 14.736176 14.561344 14.548342 12.907834 16.012190 16.201206
#>  [57] 16.205167 16.088081 16.024057 15.991279 15.954276 15.825581 16.101120
#>  [64] 15.699126 15.922914 15.901647 15.902547 15.173592 15.087845 15.088727
#>  [71] 15.089577 15.090399 15.091162 15.057795 15.282175 15.283118 15.177291
#>  [78] 15.177672 15.178054 15.178436 15.178817 15.132555 15.132873 15.133192
#>  [85] 15.133512 15.133831 15.134150 15.134468 15.134784 15.135098 15.135410
#>  [92] 15.135720 15.119379 15.119434 15.119495 15.119560 15.119630 15.119600
#>  [99] 15.150429 15.151038
#> 
#> $cvse
#>   [1] 12.130335 12.118989 11.063013  9.637070  8.448725  7.453935  6.615393
#>   [8]  5.902180  5.289366  4.757499  4.292024  3.882644  3.522694  3.208535
#>  [15]  2.938951  2.714868  2.534195  2.379993  2.317474  2.317701  2.321033
#>  [22]  2.324139  2.327396  2.330890  2.293571  2.255737  2.201555  2.155864
#>  [29]  2.119115  2.089982  2.067359  2.050134  2.037489  2.028674  2.039223
#>  [36]  2.042687  2.044122  2.046209  2.106490  2.108978  2.099147  2.324973
#>  [43]  2.322743  2.316621  2.312868  2.335301  2.324129  3.252563  3.283476
#>  [50]  3.278996  3.798130  3.981373  3.977458  3.516704  5.833418  5.830280
#>  [57]  5.831534  5.812048  5.817112  5.801180  5.806629  5.816844  5.900781
#>  [64]  5.858921  5.939096  5.942865  5.943570  5.883055  5.887645  5.888303
#>  [71]  5.888931  5.889530  5.890103  5.892107  5.932778  5.933352  5.900045
#>  [78]  5.900426  5.900792  5.901143  5.901481  5.907544  5.907863  5.908170
#>  [85]  5.908464  5.908746  5.909018  5.909278  5.909529  5.909769  5.910000
#>  [92]  5.910222  5.911314  5.911521  5.911719  5.911909  5.912092  5.912178
#>  [99]  5.939071  5.939697
#> 
#> $fold
#>  [1] 10  6  1  9  4  2  9  1  8  7  4  7  6 10  5  3  8  5  3  2  1
#> 
#> $lambda
#>   [1] 5.370736715 5.008767036 4.671192901 4.356370134 4.062765367 3.788948579
#>   [7] 3.533586126 3.295434247 3.073333009 2.866200650 2.673028319 2.492875157
#>  [13] 2.324863714 2.168175681 2.022047897 1.885768636 1.758674141 1.640145390
#>  [19] 1.529605080 1.426514817 1.330372493 1.240709840 1.157090149 1.079106146
#>  [25] 1.006378003 0.938551494 0.875296264 0.816304226 0.761288053 0.709979786
#>  [31] 0.662129525 0.617504211 0.575886494 0.537073672 0.500876704 0.467119290
#>  [37] 0.435637013 0.406276537 0.378894858 0.353358613 0.329543425 0.307333301
#>  [43] 0.286620065 0.267302832 0.249287516 0.232486371 0.216817568 0.202204790
#>  [49] 0.188576864 0.175867414 0.164014539 0.152960508 0.142651482 0.133037249
#>  [55] 0.124070984 0.115709015 0.107910615 0.100637800 0.093855149 0.087529626
#>  [61] 0.081630423 0.076128805 0.070997979 0.066212952 0.061750421 0.057588649
#>  [67] 0.053707367 0.050087670 0.046711929 0.043563701 0.040627654 0.037889486
#>  [73] 0.035335861 0.032954342 0.030733330 0.028662006 0.026730283 0.024928752
#>  [79] 0.023248637 0.021681757 0.020220479 0.018857686 0.017586741 0.016401454
#>  [85] 0.015296051 0.014265148 0.013303725 0.012407098 0.011570901 0.010791061
#>  [91] 0.010063780 0.009385515 0.008752963 0.008163042 0.007612881 0.007099798
#>  [97] 0.006621295 0.006175042 0.005758865 0.005370737
#> 
#> $fit
#> $beta
#>              5.3707     5.0088    4.6712    4.3564    4.0628    3.7889
#> (Intercept) 20.0381 21.9480358 23.729253 25.390422 26.939633 28.384433
#> am           0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.0000 -0.6084182 -1.175831 -1.705002 -2.198509 -2.658755
#>                3.5336   3.2954   3.0733    2.8662    2.6730    2.4929    2.3249
#> (Intercept) 29.731858 30.98847 32.16039 33.253332 34.272609 35.223191 36.109707
#> am           0.000000  0.00000  0.00000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.00000  0.00000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.00000  0.00000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.00000  0.00000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.00000  0.00000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.00000  0.00000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.00000  0.00000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.00000  0.00000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.00000  0.00000  0.000000  0.000000  0.000000  0.000000
#> wt          -3.087982 -3.48828 -3.86160 -4.209759 -4.534454 -4.837265 -5.119668
#>                2.1682    2.0220    1.8858    1.7587    1.6401    1.5296
#> (Intercept) 36.936475 37.707521 38.426602 38.930631 38.930631 38.930631
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -5.383037 -5.628657 -5.857722 -6.018283 -6.018283 -6.018283
#>                1.4265    1.3304    1.2407    1.1571      1.0791     1.0064
#> (Intercept) 38.930631 38.930631 38.930631 38.930631 37.86326833 36.4882127
#> am           0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.05636566  0.1289801
#> vs           0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> wt          -6.018283 -6.018283 -6.018283 -6.018283 -5.99591157 -5.9670914
#>                0.9386     0.8753     0.8163     0.7613     0.7100    0.6621
#> (Intercept) 35.205690 34.0096059 32.8941334 31.8538400 30.8836589 29.978865
#> am           0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> carb         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> cyl          0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> disp         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> drat         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> gear         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> hp           0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> qsec         0.196708  0.2598712  0.3187775  0.3737137  0.4249473  0.472728
#> vs           0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> wt          -5.940211 -5.9151417 -5.8917622 -5.8699584 -5.8496241 -5.830660
#>                 0.6175     0.5759    0.5371     0.5009     0.4671     0.4356
#> (Intercept) 29.1350508 28.3481070 27.614201 26.9297570 26.2914426 25.6961485
#> am           0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> qsec         0.5172884  0.5588457  0.597602  0.6337464  0.6674547  0.6988912
#> vs           0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> wt          -5.8129746 -5.7964808 -5.781099 -5.7667533 -5.7533747 -5.7408978
#>                0.4063     0.3789     0.3534     0.3295     0.3073    0.2866
#> (Intercept) 25.140975 24.9760670 24.9760498 24.9760490 24.9760489 20.264219
#> am           0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> carb         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> cyl          0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> disp         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> drat         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> gear         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> hp           0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> qsec         0.728209  0.7369175  0.7369184  0.7369185  0.7369185  1.146396
#> vs           0.000000  0.0000000  0.0000000  0.0000000  0.0000000 -2.065731
#> wt          -5.729262 -5.7258054 -5.7258050 -5.7258050 -5.7258050 -6.219034
#>                0.2673    0.2493     0.2325     0.2168        0.2022      0.1886
#> (Intercept) 20.261003 20.258793 13.8909808 10.1462663  1.013141e+01 13.97827189
#> am           0.000000  0.000000  0.0000000  0.0000000  0.000000e+00  0.00000000
#> carb         0.000000  0.000000  0.0000000  0.0000000  0.000000e+00  0.00000000
#> cyl          0.000000  0.000000  0.0000000  0.0000000  0.000000e+00  0.00000000
#> disp         0.000000  0.000000  0.0000000  0.0000000  4.428382e-06  0.01770017
#> drat         0.000000  0.000000  0.0000000  0.0000000  0.000000e+00  0.00000000
#> gear         0.000000  0.000000  0.5547576  0.8809913  8.820510e-01  1.38948351
#> hp           0.000000  0.000000  0.0000000  0.0000000  0.000000e+00 -0.02210622
#> qsec         1.146626  1.146785  1.3579119  1.4820315  1.482568e+00  1.19357916
#> vs          -2.066511 -2.067048 -2.4224632 -2.6309794 -2.631163e+00 -1.91136109
#> wt          -6.219190 -6.219298 -5.9830609 -5.8439969 -5.843831e+00 -6.35250691
#>                  0.1759      0.1640      0.1530      0.1427      0.1330
#> (Intercept) 13.98408273 12.42153203 12.41563708 12.41019055 12.40516254
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.71389486  0.71390612  0.71391274  0.71391532
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.01770068  0.02679201  0.02679359  0.02679495  0.02679613
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         1.38934329  0.99182171  0.99195286  0.99207617  0.99219202
#> hp          -0.02211214 -0.03327163 -0.03326642 -0.03326152 -0.03325692
#> qsec         1.19327847  1.44202565  1.44234480  1.44263865  1.44290896
#> vs          -1.91092411 -1.99140365 -1.99181649 -1.99219896 -1.99255300
#> wt          -6.35231302 -7.53702664 -7.53741647 -7.53776755 -7.53808310
#>                  0.1241      0.1157      0.1079      0.1006      0.0939
#> (Intercept) 12.40052482  7.17390477  7.16954793  7.84744629  7.85191408
#> am           0.00000000  0.00000000  0.00000000  1.09091578  1.09128522
#> carb         0.71391435  0.58974735  0.58987732  0.56813264  0.56801986
#> cyl          0.00000000  0.83740645  0.83712456  0.98285990  0.98316129
#> disp         0.02679714  0.02366214  0.02366592  0.02265193  0.02264834
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.99230077  1.40819582  1.40814081  0.99902626  0.99892195
#> hp          -0.03325260 -0.03797233 -0.03796675 -0.03906096 -0.03906681
#> qsec         1.44315737  1.50436325  1.50469488  1.43990673  1.43957248
#> vs          -1.99288046 -1.53195702 -1.53255234 -1.10806586 -1.10735812
#> wt          -7.53836613 -7.86364956 -7.86408066 -7.58559323 -7.58511671
#>                 0.0875      0.0816      0.0761      0.0710      0.0662
#> (Intercept)  7.8560830 13.17237905 13.17689824 13.18099547 13.18470850
#> am           1.0916411  1.21012572  1.21041284  1.21068154  1.21093245
#> carb         0.5679178  0.66130474  0.66153219  0.66174722  0.66195060
#> cyl          0.9834475  0.92704728  0.92706603  0.92708161  0.92709374
#> disp         0.0226451  0.02293564  0.02293917  0.02294258  0.02294589
#> drat         0.0000000 -0.59745950 -0.59775274 -0.59802728 -0.59828423
#> gear         0.9988195  0.96986496  0.96960768  0.96936798  0.96914453
#> hp          -0.0390724 -0.04259070 -0.04259827 -0.04260525 -0.04261168
#> qsec         1.4392606  1.30990125  1.30978006  1.30967498  1.30958446
#> vs          -1.1066873 -0.98709821 -0.98677472 -0.98648152 -0.98621678
#> wt          -7.5846827 -7.67979845 -7.68012954 -7.68045547 -7.68077521
#>                  0.0618      0.0576      0.0537      0.0501      0.0467
#> (Intercept) 13.18807159 13.19111583 13.19386949 13.19635827 13.19860553
#> am           1.21116622  1.21138351  1.21158497  1.21177131  1.21194319
#> carb         0.66214299  0.66232503  0.66249727  0.66266025  0.66281444
#> cyl          0.92710228  0.92710716  0.92710841  0.92710612  0.92710044
#> disp         0.02294908  0.02295216  0.02295513  0.02295798  0.02296072
#> drat        -0.59852463 -0.59874944 -0.59895959 -0.59915593 -0.59933930
#> gear         0.96893616  0.96874181  0.96856054  0.96839148  0.96823383
#> hp          -0.04261760 -0.04262304 -0.04262803 -0.04263260 -0.04263679
#> qsec         1.30950708  1.30944154  1.30938664  1.30934131  1.30930453
#> vs          -0.98597869 -0.98576553 -0.98557561 -0.98540731 -0.98525906
#> wt          -7.68108788 -7.68139277 -7.68168926 -7.68197689 -7.68225528
#>                  0.0436      0.0406      0.0379      0.0353      0.0330
#> (Intercept) 13.20063256 13.20245874 13.20410174 13.20557768 13.20690130
#> am           1.21210132  1.21224638  1.21237908  1.21250010  1.21261011
#> carb         0.66296030  0.66309823  0.66322864  0.66335188  0.66346831
#> cyl          0.92709154  0.92707963  0.92706495  0.92704774  0.92702826
#> disp         0.02296335  0.02296587  0.02296827  0.02297057  0.02297277
#> drat        -0.59951044 -0.59967008 -0.59981891 -0.59995756 -0.60008666
#> gear         0.96808688  0.96794995  0.96782242  0.96770372  0.96759331
#> hp          -0.04264063 -0.04264413 -0.04264732 -0.04265023 -0.04265287
#> qsec         1.30927542  1.30925312  1.30923688  1.30922602  1.30921991
#> vs          -0.98512938 -0.98501682 -0.98492003 -0.98483770 -0.98476861
#> wt          -7.68252415 -7.68278331 -7.68303265 -7.68327213 -7.68350174
#>                  0.0307      0.0287      0.0267      0.0249      0.0232
#> (Intercept) 13.20808603 13.20914418 13.21008699 13.21092475 13.21166689
#> am           1.21270977  1.21279975  1.21288066  1.21295313  1.21301774
#> carb         0.66357826  0.66368203  0.66377992  0.66387223  0.66395921
#> cyl          0.92700674  0.92698346  0.92695864  0.92693253  0.92690535
#> disp         0.02297486  0.02297685  0.02297874  0.02298054  0.02298225
#> drat        -0.60020677 -0.60031844 -0.60042219 -0.60051851 -0.60060784
#> gear         0.96749068  0.96739536  0.96730692  0.96722492  0.96714897
#> hp          -0.04265527 -0.04265744 -0.04265941 -0.04266118 -0.04266278
#> qsec         1.30921796  1.30921967  1.30922457  1.30923221  1.30924224
#> vs          -0.98471160 -0.98466556 -0.98462946 -0.98460235 -0.98458332
#> wt          -7.68372156 -7.68393169 -7.68413228 -7.68432351 -7.68450558
#>                  0.0217     0.0202      0.0189      0.0176      0.0164
#> (Intercept) 13.21232205 13.2128981 13.21340243 13.21384157 13.21422165
#> am           1.21307506  1.2131256  1.21316998  1.21320859  1.21324194
#> carb         0.66404112  0.6641182  0.66419075  0.66425892  0.66432297
#> cyl          0.92687732  0.9268486  0.92681950  0.92679006  0.92676048
#> disp         0.02298387  0.0229854  0.02298685  0.02298822  0.02298952
#> drat        -0.60069064 -0.6007673 -0.60083826 -0.60090384 -0.60096441
#> gear         0.96707871  0.9670138  0.96695381  0.96689853  0.96684763
#> hp          -0.04266422 -0.0426655 -0.04266665 -0.04266768 -0.04266859
#> qsec         1.30925428  1.3092680  1.30928321  1.30929956  1.30931685
#> vs          -0.98457153 -0.9845662 -0.98456657 -0.98457200 -0.98458187
#> wt          -7.68467873 -7.6848432 -7.68499925 -7.68514716 -7.68528720
#>                  0.0153      0.0143      0.0133      0.0124      0.0116
#> (Intercept) 13.21454829 13.21482661 13.21506132 13.21525675 13.21541684
#> am           1.21327047  1.21329461  1.21331474  1.21333124  1.21334445
#> carb         0.66438309  0.66443950  0.66449238  0.66454193  0.66458831
#> cyl          0.92673092  0.92670150  0.92667234  0.92664354  0.92661519
#> disp         0.02299074  0.02299189  0.02299297  0.02299399  0.02299495
#> drat        -0.60102029 -0.60107180 -0.60111922 -0.60116285 -0.60120294
#> gear         0.96680081  0.96675781  0.96671838  0.96668226  0.96664924
#> hp          -0.04266940 -0.04267011 -0.04267074 -0.04267128 -0.04267176
#> qsec         1.30933488  1.30935348  1.30937247  1.30939172  1.30941111
#> vs          -0.98459561 -0.98461268 -0.98463261 -0.98465496 -0.98467933
#> wt          -7.68541968 -7.68554487 -7.68566308 -7.68577458 -7.68587968
#>                  0.0108      0.0101      0.0094      0.0088      0.0082
#> (Intercept) 13.21554523 13.21564524 13.21571989 13.21577198 13.21580405
#> am           1.21335469  1.21336227  1.21336747  1.21337055  1.21337174
#> carb         0.66463171  0.66467228  0.66471019  0.66474557  0.66477858
#> cyl          0.92658737  0.92656015  0.92653359  0.92650774  0.92648264
#> disp         0.02299585  0.02299670  0.02299749  0.02299823  0.02299893
#> drat        -0.60123974 -0.60127347 -0.60130436 -0.60133261 -0.60135842
#> gear         0.96661910  0.96659164  0.96656666  0.96654398  0.96652343
#> hp          -0.04267217 -0.04267252 -0.04267282 -0.04267307 -0.04267328
#> qsec         1.30943051  1.30944983  1.30946899  1.30948792  1.30950656
#> vs          -0.98470536 -0.98473273 -0.98476114 -0.98479033 -0.98482005
#> wt          -7.68597866 -7.68607180 -7.68615937 -7.68624164 -7.68631889
#>                  0.0076      0.0071      0.0066      0.0062      0.0058
#> (Intercept) 13.21581841 13.21581717 13.21580227 13.21577546 13.21573833
#> am           1.21337127  1.21336933  1.21336613  1.21336182  1.21335657
#> carb         0.66480936  0.66483802  0.66486470  0.66488952  0.66491259
#> cyl          0.92645831  0.92643478  0.92641208  0.92639020  0.92636916
#> disp         0.02299958  0.02300019  0.02300076  0.02300129  0.02300179
#> drat        -0.60138195 -0.60140339 -0.60142289 -0.60144060 -0.60145665
#> gear         0.96650485  0.96648809  0.96647302  0.96645948  0.96644738
#> hp          -0.04267345 -0.04267358 -0.04267369 -0.04267376 -0.04267382
#> qsec         1.30952484  1.30954273  1.30956019  1.30957718  1.30959369
#> vs          -0.98485010 -0.98488029 -0.98491046 -0.98494046 -0.98497017
#> wt          -7.68639136 -7.68645929 -7.68652293 -7.68658251 -7.68663825
#>                  0.0054
#> (Intercept) 13.21569232
#> am           1.21335051
#> carb         0.66493401
#> cyl          0.92634896
#> disp         0.02300225
#> drat        -0.60147118
#> gear         0.96643658
#> hp          -0.04267385
#> qsec         1.30960969
#> vs          -0.98499947
#> wt          -7.68669036
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1   4   3   3   3   3   3   3   3   3   3   3   3   3
#>  [37]   3   3   3   1   1   1  38   1   1  93  25   2 103   1  88   1   1   1
#>  [55]   1  76   1  73   1   1  56   1   1   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.370736715 5.008767036 4.671192901 4.356370134 4.062765367 3.788948579
#>   [7] 3.533586126 3.295434247 3.073333009 2.866200650 2.673028319 2.492875157
#>  [13] 2.324863714 2.168175681 2.022047897 1.885768636 1.758674141 1.640145390
#>  [19] 1.529605080 1.426514817 1.330372493 1.240709840 1.157090149 1.079106146
#>  [25] 1.006378003 0.938551494 0.875296264 0.816304226 0.761288053 0.709979786
#>  [31] 0.662129525 0.617504211 0.575886494 0.537073672 0.500876704 0.467119290
#>  [37] 0.435637013 0.406276537 0.378894858 0.353358613 0.329543425 0.307333301
#>  [43] 0.286620065 0.267302832 0.249287516 0.232486371 0.216817568 0.202204790
#>  [49] 0.188576864 0.175867414 0.164014539 0.152960508 0.142651482 0.133037249
#>  [55] 0.124070984 0.115709015 0.107910615 0.100637800 0.093855149 0.087529626
#>  [61] 0.081630423 0.076128805 0.070997979 0.066212952 0.061750421 0.057588649
#>  [67] 0.053707367 0.050087670 0.046711929 0.043563701 0.040627654 0.037889486
#>  [73] 0.035335861 0.032954342 0.030733330 0.028662006 0.026730283 0.024928752
#>  [79] 0.023248637 0.021681757 0.020220479 0.018857686 0.017586741 0.016401454
#>  [85] 0.015296051 0.014265148 0.013303725 0.012407098 0.011570901 0.010791061
#>  [91] 0.010063780 0.009385515 0.008752963 0.008163042 0.007612881 0.007099798
#>  [97] 0.006621295 0.006175042 0.005758865 0.005370737
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 42
#> 
#> $loss
#>   [1] 725.70952 609.42555 512.13672 431.10948 363.98381 308.72340 263.57237
#>   [8] 227.01772 197.75676 174.66880 156.79063 143.29523 133.47329 126.71723
#>  [15] 122.50731 120.39959 119.96845 119.96845 119.96845 119.96845 119.96845
#>  [22] 119.96845 119.96845 115.78384 110.88345 106.81067 103.44538 100.68352
#>  [29]  98.43536  96.62361  95.18175  94.05258  93.18695  92.54269  92.08364
#>  [36]  91.77885  91.60187  91.53010  91.52613  91.52613  91.52613  91.52613
#>  [43]  85.14220  85.14219  85.14219  82.13034  81.64847  81.64414  67.58179
#>  [50]  67.58178  63.25935  63.25934  63.25933  63.25932  63.25931  59.65570
#>  [57]  59.65568  58.49944  58.49941  58.49939  57.89108  57.89107  57.89105
#>  [64]  57.89104  57.89102  57.89101  57.89100  57.89100  57.89099  57.89098
#>  [71]  57.89098  57.89097  57.89097  57.89096  57.89096  57.89095  57.89095
#>  [78]  57.89095  57.89095  57.89095  57.89094  57.89094  57.89094  57.89094
#>  [85]  57.89094  57.89094  57.89094  57.89094  57.89093  57.89093  57.89093
#>  [92]  57.89093  57.89093  57.89093  57.89093  57.89093  57.89093  57.89093
#>  [99]  57.89093  57.89093
#> 
#> $linear.predictors
#>     5.3707   5.0088   4.6712   4.3564   4.0628   3.7889   3.5336   3.2954
#> 1  20.0381 20.19883 20.34874 20.48854 20.61892 20.74051 20.85391 20.95967
#> 2  20.0381 20.53651 21.00132 21.43482 21.83909 22.21612 22.56774 22.89566
#> 3  20.0381 19.99197 19.94896 19.90884 19.87143 19.83654 19.80400 19.77365
#> 4  20.0381 19.84291 19.66088 19.49111 19.33279 19.18514 19.04744 18.91902
#> 5  20.0381 19.77598 19.53154 19.30356 19.09096 18.89268 18.70776 18.53531
#> 6  20.0381 19.85508 19.68439 19.52521 19.37676 19.23832 19.10920 18.98879
#> 7  20.0381 19.85508 19.68439 19.52521 19.37676 19.23832 19.10920 18.98879
#> 8  20.0381 19.47177 18.94362 18.45106 17.99170 17.56330 17.16377 16.79117
#> 9  20.0381 19.67864 19.34340 19.03076 18.73920 18.46728 18.21369 17.97719
#> 10 20.0381 19.64822 19.28461 18.94551 18.62927 18.33434 18.05929 17.80277
#> 11 20.0381 18.75384 17.55614 16.43916 15.39746 14.42597 13.51995 12.67500
#> 12 20.0381 20.96544 21.83029 22.63684 23.38904 24.09054 24.74477 25.35490
#> 13 20.0381 20.83159 21.57160 22.26174 22.90537 23.50562 24.06541 24.58748
#> 14 20.0381 20.44828 20.83083 21.18759 21.52031 21.83060 22.11998 22.38986
#> 15 20.0381 19.61171 19.21406 18.84321 18.49736 18.17481 17.87401 17.59348
#> 16 20.0381 19.60867 19.20818 18.83469 18.48637 18.16152 17.85857 17.57603
#> 17 20.0381 20.64602 21.21297 21.74172 22.23482 22.69470 23.12358 23.52355
#> 18 20.0381 21.02750 21.95022 22.81075 23.61329 24.36174 25.05974 25.71070
#> 19 20.0381 20.01935 20.00187 19.98556 19.97036 19.95618 19.94296 19.93062
#> 20 20.0381 19.77598 19.53154 19.30356 19.09096 18.89268 18.70776 18.53531
#> 21 20.0381 20.25663 20.46044 20.65052 20.82778 20.99309 21.14727 21.29105
#>      3.0733   2.8662   2.6730   2.4929    2.3249    2.1682    2.0220    1.8858
#> 1  21.05829 21.15027 21.23605 21.31605 21.390663 21.460242 21.525133 21.585650
#> 2  23.20148 23.48669 23.75268 24.00074 24.232078 24.447828 24.649037 24.836686
#> 3  19.74535 19.71896 19.69434 19.67138 19.649976 19.630010 19.611389 19.594024
#> 4  18.79926 18.68756 18.58340 18.48625 18.395657 18.311166 18.232369 18.158882
#> 5  18.37448 18.22449 18.08461 17.95416 17.832494 17.719031 17.613216 17.514533
#> 6  18.87649 18.77176 18.67409 18.58300 18.498050 18.418826 18.344942 18.276037
#> 7  18.87649 18.77176 18.67409 18.58300 18.498050 18.418826 18.344942 18.276037
#> 8  16.44368 16.11961 15.81738 15.53552 15.272660 15.027513 14.798888 14.585671
#> 9  17.75663 17.55093 17.35910 17.18019 17.013347 16.857745 16.712631 16.577297
#> 10 17.56355 17.34044 17.13237 16.93833 16.757363 16.588594 16.431198 16.284411
#> 11 11.88699 11.15210 10.46673  9.82755  9.231452  8.675529  8.157073  7.673559
#> 12 25.92391 26.45457 26.94947 27.41101 27.841444 28.242870 28.617240 28.966380
#> 13 25.07436 25.52842 25.95189 26.34681 26.715117 27.058601 27.378936 27.677681
#> 14 22.64155 22.87627 23.09518 23.29933 23.489726 23.667288 23.832882 23.987316
#> 15 17.33185 17.08786 16.86031 16.64809 16.450183 16.265611 16.093479 15.932948
#> 16 17.31254 17.06681 16.83763 16.62391 16.424585 16.238696 16.065336 15.903659
#> 17 23.89657 24.24445 24.56888 24.87144 25.153618 25.416775 25.662196 25.891076
#> 18 26.31779 26.88397 27.41198 27.90441 28.363650 28.791939 29.191363 29.563868
#> 19 19.91912 19.90839 19.89839 19.88906 19.880361 19.872246 19.864679 19.857622
#> 20 18.37448 18.22449 18.08461 17.95416 17.832494 17.719031 17.613216 17.514533
#> 21 21.42515 21.55020 21.66683 21.77559 21.877031 21.971631 22.059855 22.142133
#>       1.7587    1.6401    1.5296    1.4265    1.3304    1.2407    1.1571
#> 1  21.628068 21.628068 21.628068 21.628068 21.628068 21.628068 21.628068
#> 2  24.968215 24.968215 24.968215 24.968215 24.968215 24.968215 24.968215
#> 3  19.581852 19.581852 19.581852 19.581852 19.581852 19.581852 19.581852
#> 4  18.107373 18.107373 18.107373 18.107373 18.107373 18.107373 18.107373
#> 5  17.445362 17.445362 17.445362 17.445362 17.445362 17.445362 17.445362
#> 6  18.227738 18.227738 18.227738 18.227738 18.227738 18.227738 18.227738
#> 7  18.227738 18.227738 18.227738 18.227738 18.227738 18.227738 18.227738
#> 8  14.436220 14.436220 14.436220 14.436220 14.436220 14.436220 14.436220
#> 9  16.482437 16.482437 16.482437 16.482437 16.482437 16.482437 16.482437
#> 10 16.181522 16.181522 16.181522 16.181522 16.181522 16.181522 16.181522
#> 11  7.334647  7.334647  7.334647  7.334647  7.334647  7.334647  7.334647
#> 12 29.211104 29.211104 29.211104 29.211104 29.211104 29.211104 29.211104
#> 13 27.887082 27.887082 27.887082 27.887082 27.887082 27.887082 27.887082
#> 14 24.095564 24.095564 24.095564 24.095564 24.095564 24.095564 24.095564
#> 15 15.820425 15.820425 15.820425 15.820425 15.820425 15.820425 15.820425
#> 16 15.790334 15.790334 15.790334 15.790334 15.790334 15.790334 15.790334
#> 17 26.051506 26.051506 26.051506 26.051506 26.051506 26.051506 26.051506
#> 18 29.824969 29.824969 29.824969 29.824969 29.824969 29.824969 29.824969
#> 19 19.852675 19.852675 19.852675 19.852675 19.852675 19.852675 19.852675
#> 20 17.445362 17.445362 17.445362 17.445362 17.445362 17.445362 17.445362
#> 21 22.199805 22.199805 22.199805 22.199805 22.199805 22.199805 22.199805
#>       1.0791    1.0064    0.9386    0.8753    0.8163   0.7613    0.7100
#> 1  21.584366 21.528066 21.475554 21.426582 21.380910 21.33832 21.298593
#> 2  25.001718 25.044880 25.085137 25.122681 25.157694 25.19035 25.220801
#> 3  19.682161 19.811387 19.931916 20.044322 20.149152 20.24692 20.338093
#> 4  18.257128 18.450054 18.629997 18.797812 18.954317 19.10027 19.236395
#> 5  17.350696 17.228741 17.114992 17.008910 16.909978 16.81771 16.731667
#> 6  18.268824 18.321754 18.371121 18.417162 18.460100 18.50014 18.537488
#> 7  18.302644 18.399142 18.489146 18.573085 18.651366 18.72437 18.792457
#> 8  14.440671 14.446404 14.451752 14.456739 14.461390 14.46573 14.469772
#> 9  16.490554 16.501011 16.510765 16.519861 16.528344 16.53626 16.543634
#> 10 16.213305 16.254249 16.292437 16.328052 16.361267 16.39224 16.421132
#> 11  7.398187  7.480045  7.556394  7.627597  7.694001  7.75593  7.813685
#> 12 29.223763 29.240071 29.255282 29.269467 29.282697 29.29503 29.306541
#> 13 27.982447 28.105304 28.219892 28.326758 28.426422 28.51937 28.606051
#> 14 24.211223 24.360224 24.499198 24.628805 24.749677 24.86240 24.967532
#> 15 15.707563 15.562165 15.426551 15.300077 15.182128 15.07213 14.969541
#> 16 15.770023 15.743857 15.719451 15.696691 15.675464 15.65567 15.637206
#> 17 25.973324 25.872604 25.778663 25.691052 25.609346 25.53315 25.462084
#> 18 29.744034 29.639767 29.542516 29.451820 29.367237 29.28835 29.214788
#> 19 19.673531 19.442744 19.227488 19.026740 18.839521 18.66492 18.502087
#> 20 17.280803 17.068806 16.871075 16.686670 16.514694 16.35431 16.204732
#> 21 22.243035 22.298728 22.350673 22.399117 22.444296 22.48643 22.525724
#>       0.6621   0.6175    0.5759    0.5371    0.5009    0.4671    0.4356
#> 1  21.261547 21.22700 21.194778 21.164729 21.136705 21.110570 21.086196
#> 2  25.249201 25.27569 25.300389 25.323426 25.344909 25.364945 25.383631
#> 3  20.423125 20.50242 20.576381 20.645352 20.709675 20.769663 20.825607
#> 4  19.363341 19.48173 19.592142 19.695112 19.791142 19.880700 19.964223
#> 5  16.651419 16.57658 16.506786 16.441695 16.380990 16.324378 16.271580
#> 6  18.572316 18.60480 18.635088 18.663338 18.689684 18.714255 18.737169
#> 7  18.855953 18.91517 18.970396 19.021900 19.069932 19.114728 19.156504
#> 8  14.473545 14.47706 14.480344 14.483404 14.486258 14.488920 14.491402
#> 9  16.550515 16.55693 16.562917 16.568498 16.573704 16.578558 16.583085
#> 10 16.448073 16.47320 16.496631 16.518484 16.538864 16.557871 16.575597
#> 11  7.867548  7.91778  7.964627  8.008317  8.049062  8.087061  8.122499
#> 12 29.317271 29.32728 29.336612 29.345316 29.353433 29.361004 29.368064
#> 13 28.686891 28.76228 28.832593 28.898165 28.959318 29.016349 29.069536
#> 14 25.065575 25.15701 25.242283 25.321809 25.395975 25.465143 25.529649
#> 15 14.873868 14.78464 14.701432 14.623829 14.551456 14.483961 14.421015
#> 16 15.619989 15.60393 15.588957 15.574991 15.561966 15.549820 15.538492
#> 17 25.395810 25.33400 25.276360 25.222603 25.172470 25.125714 25.082111
#> 18 29.146179 29.08219 29.022523 28.966873 28.914973 28.866571 28.821432
#> 19 18.350228 18.20860 18.076525 17.953347 17.838472 17.731338 17.631425
#> 20 16.065237 15.93514 15.813817 15.700668 15.595145 15.496734 15.404955
#> 21 22.562370 22.59655 22.628419 22.658144 22.685866 22.711719 22.735829
#>       0.4063    0.3789    0.3534    0.3295    0.3073    0.2866    0.2673
#> 1  21.063465 21.056713 21.056712 21.056712 21.056712 21.896155 21.896413
#> 2  25.401057 25.406234 25.406234 25.406234 25.406234 25.104758 25.104689
#> 3  20.877782 20.893279 20.893281 20.893281 20.893281 20.490231 20.490213
#> 4  20.042115 20.065253 20.065255 20.065255 20.065255 19.860756 19.860880
#> 5  16.222341 16.207715 16.207714 16.207714 16.207714 16.221179 16.221057
#> 6  18.758539 18.764887 18.764888 18.764888 18.764888 17.784057 17.783741
#> 7  19.195465 19.207038 19.207039 19.207039 19.207039 18.471894 18.471717
#> 8  14.493716 14.494404 14.494404 14.494404 14.494404 14.900040 14.900199
#> 9  16.587307 16.588561 16.588562 16.588562 16.588562 17.243791 17.244049
#> 10 16.592128 16.597038 16.597039 16.597039 16.597039 17.391397 17.391740
#> 11  8.155549  8.165366  8.165367  8.165367  8.165367  8.226489  8.226598
#> 12 29.374648 29.376604 29.376604 29.376604 29.376604 29.386001 29.386022
#> 13 29.119139 29.133873 29.133874 29.133875 29.133875 29.599840 29.600144
#> 14 25.589807 25.607676 25.607678 25.607678 25.607678 25.807952 25.808183
#> 15 14.362311 14.344873 14.344872 14.344871 14.344871 14.049090 14.048826
#> 16 15.527927 15.524789 15.524789 15.524789 15.524789 15.898084 15.898198
#> 17 25.041445 25.029366 25.029365 25.029365 25.029365 26.100298 26.100598
#> 18 28.779334 28.766830 28.766828 28.766828 28.766828 28.163181 28.162844
#> 19 17.538246 17.510568 17.510565 17.510565 17.510565 17.172622 17.172253
#> 20 15.319362 15.293938 15.293935 15.293935 15.293935 14.799648 14.799240
#> 21 22.758315 22.764994 22.764995 22.764995 22.764995 22.232538 22.232395
#>       0.2493   0.2325    0.2168    0.2022   0.1886   0.1759   0.1640   0.1530
#> 1  21.896591 22.02037 22.092916 22.092609 21.98781 21.98793 22.74557 22.74538
#> 2  25.104642 25.07759 25.061785 25.061825 24.95528 24.95532 24.26078 24.26070
#> 3  20.490202 20.29506 20.180502 20.180741 21.15019 21.15028 21.17339 21.17341
#> 4  19.860966 19.88838 19.904708 19.905259 20.05124 20.05115 19.73382 19.73392
#> 5  16.220973 15.70505 15.401550 15.400552 15.33060 15.33066 15.68065 15.68070
#> 6  17.783525 17.95561 18.057078 18.057402 17.86220 17.86241 18.11262 18.11229
#> 7  18.471596 18.77035 18.946297 18.946943 18.57835 18.57838 18.97784 18.97770
#> 8  14.900308 14.83186 14.791520 14.791069 13.96288 13.96291 13.35457 13.35444
#> 9  17.244227 17.13769 17.074885 17.074485 16.36145 16.36135 16.20557 16.20563
#> 10 17.391976 17.38170 17.375498 17.375320 16.52126 16.52104 16.40553 16.40570
#> 11  8.226672  8.55944  8.755182  8.756107 10.07932 10.07935 10.43595 10.43600
#> 12 29.386036 29.17343 29.048420 29.048152 29.66101 29.66117 30.65725 30.65716
#> 13 29.600353 29.73108 29.807944 29.808432 29.54180 29.54151 29.71943 29.71975
#> 14 25.808342 25.55636 25.408258 25.408067 24.44144 24.44121 24.38602 24.38624
#> 15 14.048645 13.50572 13.186397 13.185169 12.92519 12.92542 12.75766 12.75746
#> 16 15.898276 15.70278 15.587709 15.587582 17.28334 17.28351 17.32573 17.32573
#> 17 26.100804 26.53815 26.794995 26.795278 27.38176 27.38179 26.95638 26.95632
#> 18 28.162613 28.13865 28.124608 28.124600 28.75976 28.75989 28.57196 28.57188
#> 19 17.172000 17.38819 17.515209 17.515506 18.47186 18.47184 17.87350 17.87363
#> 20 14.798959 15.13075 15.325813 15.326009 13.59566 13.59525 14.15659 14.15693
#> 21 22.232297 22.31180 22.358726 22.358895 21.89759 21.89764 21.30917 21.30902
#>      0.1427   0.1330   0.1241   0.1157    0.1079    0.1006    0.0939    0.0875
#> 1  22.74520 22.74503 22.74488 22.79537 22.795248 23.130392 23.130633 23.130863
#> 2  24.26063 24.26056 24.26051 23.99072 23.990651 24.337800 24.337975 24.338136
#> 3  21.17344 21.17345 21.17347 21.37179 21.371758 21.353347 21.353372 21.353400
#> 4  19.73402 19.73411 19.73419 20.02761 20.027578 20.065795 20.065823 20.065849
#> 5  15.68075 15.68078 15.68081 15.42777 15.427954 15.292248 15.292046 15.291861
#> 6  18.11198 18.11170 18.11144 18.43223 18.431793 18.152993 18.153310 18.153604
#> 7  18.97756 18.97744 18.97733 19.33485 19.334610 19.016937 19.017054 19.017160
#> 8  13.35433 13.35423 13.35414 13.72885 13.728529 13.809243 13.809553 13.809837
#> 9  16.20570 16.20576 16.20582 16.70337 16.703255 16.676326 16.676407 16.676481
#> 10 16.40587 16.40602 16.40616 16.91193 16.911929 16.873009 16.872980 16.872952
#> 11 10.43604 10.43607 10.43610  9.60523  9.605599  9.729306  9.729022  9.728759
#> 12 30.65707 30.65699 30.65690 30.78153 30.781510 30.994027 30.994138 30.994253
#> 13 29.72004 29.72030 29.72055 29.93531 29.935583 30.132143 30.131952 30.131776
#> 14 24.38644 24.38663 24.38680 23.68283 23.683282 23.281661 23.281105 23.280577
#> 15 12.75727 12.75709 12.75693 12.42109 12.420974 12.398458 12.398565 12.398664
#> 16 17.32573 17.32573 17.32572 17.51060 17.510567 17.452576 17.452592 17.452615
#> 17 26.95627 26.95623 26.95619 26.42973 26.429813 25.984950 25.984720 25.984495
#> 18 28.57180 28.57172 28.57165 28.69748 28.697329 28.490863 28.490937 28.491007
#> 19 17.87375 17.87386 17.87396 18.43934 18.439216 18.539953 18.540090 18.540224
#> 20 14.15724 14.15753 14.15779 13.92416 13.924627 14.016312 14.015920 14.015553
#> 21 21.30889 21.30877 21.30867 20.64820 20.648194 21.071660 21.071805 21.071935
#>       0.0816    0.0761    0.0710    0.0662    0.0618    0.0576    0.0537
#> 1  23.339198 23.339572 23.339915 23.340232 23.340522 23.340790 23.341036
#> 2  24.420384 24.420312 24.420241 24.420172 24.420103 24.420037 24.419971
#> 3  21.484634 21.484799 21.484957 21.485107 21.485250 21.485386 21.485514
#> 4  20.272071 20.272075 20.272081 20.272088 20.272095 20.272102 20.272109
#> 5  15.379788 15.379968 15.380142 15.380311 15.380473 15.380630 15.380781
#> 6  18.088244 18.088234 18.088214 18.088184 18.088146 18.088101 18.088051
#> 7  18.874185 18.874102 18.874019 18.873934 18.873850 18.873766 18.873683
#> 8  13.842889 13.842723 13.842560 13.842399 13.842242 13.842087 13.841937
#> 9  16.716001 16.715924 16.715850 16.715780 16.715713 16.715649 16.715588
#> 10 16.855972 16.855829 16.855697 16.855575 16.855461 16.855356 16.855258
#> 11  9.720623  9.720768  9.720909  9.721046  9.721179  9.721309  9.721435
#> 12 30.738197 30.738477 30.738739 30.738986 30.739217 30.739434 30.739637
#> 13 29.960014 29.959919 29.959838 29.959770 29.959712 29.959664 29.959625
#> 14 23.157462 23.157199 23.156959 23.156742 23.156545 23.156369 23.156211
#> 15 12.202949 12.202905 12.202859 12.202813 12.202768 12.202723 12.202679
#> 16 17.506659 17.506856 17.507044 17.507224 17.507395 17.507558 17.507712
#> 17 25.939868 25.939622 25.939389 25.939170 25.938963 25.938769 25.938588
#> 18 28.909333 28.909532 28.909717 28.909890 28.910050 28.910199 28.910336
#> 19 18.227217 18.226993 18.226787 18.226597 18.226421 18.226259 18.226107
#> 20 14.167057 14.167084 14.167117 14.167157 14.167202 14.167251 14.167305
#> 21 20.997255 20.997108 20.996965 20.996825 20.996691 20.996561 20.996437
#>       0.0501    0.0467    0.0436    0.0406    0.0379    0.0353    0.0330
#> 1  23.341261 23.341469 23.341659 23.341833 23.341992 23.342138 23.342272
#> 2  24.419908 24.419846 24.419786 24.419728 24.419672 24.419617 24.419564
#> 3  21.485635 21.485748 21.485855 21.485956 21.486049 21.486137 21.486219
#> 4  20.272116 20.272122 20.272128 20.272133 20.272137 20.272140 20.272143
#> 5  15.380927 15.381067 15.381202 15.381331 15.381455 15.381574 15.381688
#> 6  18.087996 18.087938 18.087877 18.087815 18.087750 18.087685 18.087620
#> 7  18.873601 18.873521 18.873443 18.873366 18.873293 18.873221 18.873152
#> 8  13.841790 13.841646 13.841507 13.841372 13.841242 13.841115 13.840993
#> 9  16.715530 16.715474 16.715420 16.715369 16.715320 16.715273 16.715228
#> 10 16.855168 16.855083 16.855004 16.854931 16.854863 16.854800 16.854741
#> 11  9.721558  9.721678  9.721794  9.721906  9.722016  9.722122  9.722224
#> 12 30.739827 30.740004 30.740169 30.740324 30.740467 30.740601 30.740725
#> 13 29.959593 29.959567 29.959547 29.959532 29.959521 29.959513 29.959509
#> 14 23.156071 23.155947 23.155838 23.155743 23.155662 23.155592 23.155533
#> 15 12.202635 12.202593 12.202553 12.202514 12.202476 12.202440 12.202405
#> 16 17.507858 17.507995 17.508124 17.508246 17.508360 17.508467 17.508567
#> 17 25.938419 25.938261 25.938115 25.937980 25.937855 25.937741 25.937635
#> 18 28.910464 28.910582 28.910690 28.910791 28.910883 28.910968 28.911046
#> 19 18.225966 18.225834 18.225710 18.225593 18.225484 18.225381 18.225283
#> 20 14.167362 14.167422 14.167484 14.167548 14.167613 14.167679 14.167745
#> 21 20.996317 20.996203 20.996093 20.995989 20.995890 20.995796 20.995707
#>       0.0307    0.0287    0.0267   0.0249    0.0232    0.0217    0.0202
#> 1  23.342394 23.342505 23.342606 23.34270 23.342782 23.342859 23.342928
#> 2  24.419513 24.419464 24.419417 24.41937 24.419327 24.419285 24.419245
#> 3  21.486295 21.486366 21.486431 21.48649 21.486548 21.486600 21.486648
#> 4  20.272145 20.272147 20.272147 20.27215 20.272147 20.272146 20.272145
#> 5  15.381796 15.381900 15.381999 15.38209 15.382183 15.382269 15.382350
#> 6  18.087555 18.087490 18.087427 18.08736 18.087303 18.087243 18.087185
#> 7  18.873086 18.873022 18.872961 18.87290 18.872848 18.872795 18.872745
#> 8  13.840876 13.840763 13.840654 13.84055 13.840450 13.840354 13.840263
#> 9  16.715185 16.715143 16.715104 16.71507 16.715030 16.714996 16.714963
#> 10 16.854686 16.854635 16.854587 16.85454 16.854502 16.854463 16.854428
#> 11  9.722323  9.722419  9.722511  9.72260  9.722686  9.722768  9.722847
#> 12 30.740840 30.740947 30.741046 30.74114 30.741222 30.741300 30.741372
#> 13 29.959508 29.959509 29.959512 29.95952 29.959523 29.959530 29.959538
#> 14 23.155484 23.155444 23.155413 23.15539 23.155372 23.155360 23.155355
#> 15 12.202372 12.202340 12.202310 12.20228 12.202254 12.202228 12.202204
#> 16 17.508660 17.508747 17.508828 17.50890 17.508973 17.509037 17.509097
#> 17 25.937539 25.937451 25.937370 25.93730 25.937232 25.937172 25.937119
#> 18 28.911118 28.911184 28.911244 28.91130 28.911349 28.911395 28.911436
#> 19 18.225191 18.225104 18.225022 18.22494 18.224870 18.224800 18.224733
#> 20 14.167811 14.167877 14.167943 14.16801 14.168072 14.168135 14.168197
#> 21 20.995622 20.995542 20.995467 20.99540 20.995328 20.995265 20.995205
#>       0.0189    0.0176    0.0164   0.0153    0.0143    0.0133   0.0124
#> 1  23.342990 23.343046 23.343097 23.34314 23.343184 23.343221 23.34325
#> 2  24.419206 24.419169 24.419133 24.41910 24.419067 24.419035 24.41901
#> 3  21.486692 21.486732 21.486769 21.48680 21.486834 21.486863 21.48689
#> 4  20.272143 20.272141 20.272139 20.27214 20.272133 20.272130 20.27213
#> 5  15.382427 15.382500 15.382569 15.38263 15.382696 15.382754 15.38281
#> 6  18.087128 18.087074 18.087021 18.08697 18.086922 18.086875 18.08683
#> 7  18.872698 18.872653 18.872611 18.87257 18.872534 18.872499 18.87247
#> 8  13.840175 13.840092 13.840013 13.83994 13.839866 13.839798 13.83973
#> 9  16.714932 16.714902 16.714874 16.71485 16.714822 16.714797 16.71477
#> 10 16.854395 16.854364 16.854336 16.85431 16.854286 16.854263 16.85424
#> 11  9.722922  9.722995  9.723064  9.72313  9.723193  9.723253  9.72331
#> 12 30.741438 30.741499 30.741554 30.74161 30.741652 30.741695 30.74173
#> 13 29.959547 29.959557 29.959567 29.95958 29.959588 29.959598 29.95961
#> 14 23.155354 23.155357 23.155364 23.15537 23.155387 23.155403 23.15542
#> 15 12.202181 12.202159 12.202138 12.20212 12.202100 12.202083 12.20207
#> 16 17.509152 17.509203 17.509250 17.50929 17.509332 17.509368 17.50940
#> 17 25.937071 25.937029 25.936991 25.93696 25.936928 25.936903 25.93688
#> 18 28.911474 28.911508 28.911539 28.91157 28.911592 28.911615 28.91164
#> 19 18.224670 18.224610 18.224553 18.22450 18.224449 18.224400 18.22435
#> 20 14.168257 14.168315 14.168372 14.16843 14.168481 14.168532 14.16858
#> 21 20.995149 20.995096 20.995046 20.99500 20.994956 20.994915 20.99488
#>       0.0116    0.0108    0.0101    0.0094    0.0088    0.0082    0.0076
#> 1  23.343283 23.343309 23.343332 23.343352 23.343370 23.343385 23.343398
#> 2  24.418978 24.418951 24.418925 24.418901 24.418878 24.418856 24.418835
#> 3  21.486912 21.486933 21.486953 21.486970 21.486986 21.487001 21.487013
#> 4  20.272123 20.272119 20.272115 20.272111 20.272108 20.272104 20.272100
#> 5  15.382861 15.382909 15.382955 15.382998 15.383039 15.383077 15.383112
#> 6  18.086788 18.086747 18.086709 18.086672 18.086637 18.086604 18.086572
#> 7  18.872435 18.872406 18.872378 18.872353 18.872329 18.872307 18.872287
#> 8  13.839672 13.839614 13.839559 13.839507 13.839458 13.839412 13.839369
#> 9  16.714753 16.714733 16.714713 16.714695 16.714678 16.714662 16.714647
#> 10 16.854223 16.854206 16.854190 16.854175 16.854161 16.854149 16.854137
#> 11  9.723364  9.723416  9.723465  9.723511  9.723555  9.723597  9.723636
#> 12 30.741770 30.741803 30.741832 30.741859 30.741883 30.741905 30.741925
#> 13 29.959620 29.959631 29.959641 29.959652 29.959662 29.959672 29.959682
#> 14 23.155439 23.155460 23.155481 23.155504 23.155527 23.155551 23.155575
#> 15 12.202051 12.202037 12.202023 12.202010 12.201998 12.201987 12.201976
#> 16 17.509432 17.509460 17.509485 17.509508 17.509529 17.509548 17.509565
#> 17 25.936861 25.936845 25.936831 25.936820 25.936811 25.936803 25.936797
#> 18 28.911653 28.911669 28.911684 28.911697 28.911708 28.911718 28.911726
#> 19 18.224312 18.224271 18.224232 18.224196 18.224162 18.224129 18.224099
#> 20 14.168630 14.168676 14.168720 14.168762 14.168803 14.168841 14.168878
#> 21 20.994841 20.994807 20.994775 20.994746 20.994719 20.994693 20.994669
#>       0.0071    0.0066    0.0062    0.0058    0.0054
#> 1  23.343410 23.343420 23.343428 23.343435 23.343441
#> 2  24.418815 24.418796 24.418778 24.418762 24.418746
#> 3  21.487025 21.487035 21.487045 21.487053 21.487060
#> 4  20.272096 20.272093 20.272089 20.272086 20.272082
#> 5  15.383146 15.383177 15.383206 15.383233 15.383259
#> 6  18.086542 18.086514 18.086488 18.086462 18.086439
#> 7  18.872268 18.872250 18.872234 18.872219 18.872205
#> 8  13.839328 13.839289 13.839253 13.839219 13.839187
#> 9  16.714632 16.714619 16.714606 16.714595 16.714584
#> 10 16.854127 16.854117 16.854108 16.854100 16.854093
#> 11  9.723673  9.723708  9.723741  9.723772  9.723802
#> 12 30.741943 30.741959 30.741973 30.741986 30.741998
#> 13 29.959692 29.959701 29.959710 29.959719 29.959727
#> 14 23.155599 23.155623 23.155647 23.155671 23.155695
#> 15 12.201967 12.201957 12.201948 12.201940 12.201932
#> 16 17.509581 17.509595 17.509608 17.509619 17.509629
#> 17 25.936793 25.936790 25.936788 25.936787 25.936788
#> 18 28.911734 28.911740 28.911746 28.911751 28.911755
#> 19 18.224070 18.224043 18.224018 18.223994 18.223972
#> 20 14.168913 14.168947 14.168979 14.169009 14.169038
#> 21 20.994647 20.994626 20.994606 20.994588 20.994571
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 22.8 21.4 18.1 14.3 19.2 17.8 16.4 17.3 15.2 10.4 30.4 33.9 21.5 13.3
#> [16] 19.2 26.0 30.4 15.8 15.0 21.4
#> 
#> $X
#>               am        carb        cyl        disp       drat       gear
#>  [1,]  1.1547005  0.69631062 -0.1097643 -0.57766547  0.4594346  0.3735437
#>  [2,]  1.1547005 -1.13150476 -1.2622890 -1.03119247  0.3699343  0.3735437
#>  [3,] -0.8660254 -1.13150476 -0.1097643  0.27705848 -1.0083694 -0.9338592
#>  [4,] -0.8660254 -1.13150476 -0.1097643 -0.01075673 -1.5811710 -0.9338592
#>  [5,] -0.8660254  0.69631062  1.0427605  1.16666912 -0.7756688 -0.9338592
#>  [6,] -0.8660254  0.69631062 -0.1097643 -0.51138076  0.4952347  0.3735437
#>  [7,] -0.8660254  0.69631062 -0.1097643 -0.51138076  0.4952347  0.3735437
#>  [8,] -0.8660254  0.08703883  1.0427605  0.43230426 -1.0262695 -0.9338592
#>  [9,] -0.8660254  0.08703883  1.0427605  0.43230426 -1.0262695 -0.9338592
#> [10,] -0.8660254  0.08703883  1.0427605  0.43230426 -1.0262695 -0.9338592
#> [11,] -0.8660254  0.69631062  1.0427605  2.14349650 -1.2768702 -0.9338592
#> [12,]  1.1547005 -0.52223297 -1.2622890 -1.31290251  2.3031396  0.3735437
#> [13,]  1.1547005 -1.13150476 -1.2622890 -1.35302220  1.0322362  0.3735437
#> [14,] -0.8660254 -1.13150476 -1.2622890 -0.92566023  0.1014336 -0.9338592
#> [15,] -0.8660254  0.69631062  1.0427605  1.07945239  0.1551338 -0.9338592
#> [16,] -0.8660254 -0.52223297  1.0427605  1.51553604 -1.0083694 -0.9338592
#> [17,]  1.1547005 -0.52223297 -1.2622890 -0.92391589  1.4081372  1.6809466
#> [18,]  1.1547005 -0.52223297 -1.2622890 -1.14370205  0.2267340  1.6809466
#> [19,]  1.1547005  0.69631062  1.0427605  1.08817406  1.0322362  1.6809466
#> [20,]  1.1547005  3.13339781  1.0427605  0.65209042 -0.1849672  1.6809466
#> [21,]  1.1547005 -0.52223297 -1.2622890 -0.91781072  0.8353356  0.3735437
#>               hp        qsec         vs          wt
#>  [1,] -0.5920635 -0.41426037 -0.9534626 -0.29604374
#>  [2,] -0.8295541  0.56813691  1.0488088 -0.91795970
#>  [3,] -0.5920635  1.08096065  1.0488088  0.08494983
#>  [4,] -0.6619137  1.56289139  1.0488088  0.35948930
#>  [5,]  1.2938915 -1.14333508 -0.9534626  0.48275193
#>  [6,] -0.4104530  0.37660034  1.0488088  0.33707792
#>  [7,] -0.4104530  0.74731629  1.0488088  0.33707792
#>  [8,]  0.3858391 -0.17947360 -0.9534626  1.04303658
#>  [9,]  0.3858391 -0.05590161 -0.9534626  0.66204302
#> [10,]  0.3858391  0.19124236 -0.9534626  0.71807148
#> [11,]  0.7350900  0.17888516 -0.9534626  2.36530836
#> [12,] -1.4023256  0.51252952  1.0488088 -1.70796106
#> [13,] -1.2207151  1.36517622  1.0488088 -1.46143581
#> [14,] -0.7736739  1.43314081  1.0488088 -0.75547715
#> [15,]  1.2938915 -1.40901485 -0.9534626  0.78530564
#> [16,]  0.3159889 -0.39572457 -0.9534626  0.79090849
#> [17,] -0.8574942 -0.61197555 -0.9534626 -1.11966217
#> [18,] -0.5501534 -0.48840356  1.0488088 -1.82225913
#> [19,]  1.5593222 -1.97126738 -0.9534626  0.03452421
#> [20,]  2.5511948 -1.90948139 -0.9534626  0.48275193
#> [21,] -0.6060335  0.56195831  1.0488088 -0.40249782
#> attr(,"center")
#>  [1]   0.4285714   2.8571429   6.1904762 226.2333333   3.6433333   3.7142857
#>  [7] 152.3809524  17.6904762   0.4761905   3.1391905
#> attr(,"scale")
#>  [1]   0.4948717   1.6413036   1.7353207 114.6569016   0.5586577   0.7648752
#>  [7]  71.5817722   1.6184898   0.4994328   0.8924035
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 38
#> 
#> $lambda.min
#> [1] 0.4062765
#> 
#> $null.dev
#> [1] 34.5576
#> 
#> $Bias
#> [1] 2.122009
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 11.25862