Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 41.534233 39.263941 35.875281 31.934241 28.647846 25.921304 23.672811
#>   [8] 21.831816 20.337516 19.137550 18.186860 17.446712 16.883838 16.469698
#>  [15] 16.179835 15.867919 15.390727 14.996423 14.907023 14.907023 14.897785
#>  [22] 14.591050 14.346742 14.167930 15.635723 15.636440 14.982441 13.603163
#>  [29] 14.469437 14.453388 14.375204 14.309423 12.579121 12.360225 11.919544
#>  [36] 11.514652 11.449191 11.127575 11.084140 11.969415 11.435884 10.712814
#>  [43] 10.245143  9.726500 10.016795 10.016490  9.604878  9.275789  9.628140
#>  [50]  9.628302  9.568811 14.227963 13.837528 13.661944 13.661327 14.786064
#>  [57] 15.849070 16.787370 16.787291 17.651094 18.169163 24.525956 24.529097
#>  [64] 24.532147 24.591943 24.594820 24.824469 24.858727 24.861528 23.631657
#>  [71] 23.634037 23.636347 23.638590 23.753344 23.797521 23.799702 25.384366
#>  [78] 25.386822 25.389206 24.917751 20.985921 20.985818 20.985702 20.985869
#>  [85] 21.012285 21.012060 21.011821 21.011568 21.011303 21.011027 21.010741
#>  [92] 21.010447 21.010144 21.009835 21.009518 21.009196 21.008870 21.008538
#>  [99] 21.008203 21.007864
#> 
#> $cvse
#>   [1] 14.067817 13.920015 13.135977 11.803349 10.613953  9.551028  8.601013
#>   [8]  7.753116  6.998938  6.332140  5.748124  5.243676  4.816541  4.464884
#>  [15]  4.186636  3.957156  3.750228  3.546230  3.491082  3.491082  3.488484
#>  [22]  3.444139  3.416108  3.401120  3.892047  3.853469  3.811402  3.770168
#>  [29]  4.530908  4.594100  4.586368  4.579517  4.517721  4.526708  4.324239
#>  [36]  4.085490  4.070871  3.872773  3.833661  3.726729  3.679271  3.687382
#>  [43]  3.791996  3.764389  3.740088  3.740107  3.375617  3.350898  3.678903
#>  [50]  3.679150  3.686364  5.590708  5.412078  5.388556  5.388159  5.297357
#>  [57]  5.211644  5.243848  5.243515  5.310652  5.271597  9.412905  9.414625
#>  [64]  9.416293  9.413996  9.415564  9.482313  9.483332  9.484777  9.552225
#>  [71]  9.553613  9.554960  9.556268  9.551445  9.550125  9.551314  9.512180
#>  [78]  9.513287  9.514362  9.383016  6.180410  6.179824  6.179243  6.178628
#>  [85]  6.174405  6.173851  6.173304  6.172763  6.172228  6.171699  6.171176
#>  [92]  6.170659  6.170148  6.169642  6.169141  6.168646  6.168156  6.167671
#>  [99]  6.167191  6.166716
#> 
#> $fold
#>  [1]  9  9  1  1  6  5  4  2  8  1  3 10  2  4  3  7 10  7  6  8  5
#> 
#> $lambda
#>   [1] 5.572914137 5.197318376 4.847036512 4.520362474 4.215705172 3.931580753
#>   [7] 3.666605369 3.419488439 3.189026363 2.974096659 2.773652498 2.586717603
#>  [13] 2.412381494 2.249795055 2.098166398 1.956757005 1.824878132 1.701887454
#>  [19] 1.587185935 1.480214916 1.380453384 1.287415446 1.200647954 1.119728300
#>  [25] 1.044262361 0.973882573 0.908246147 0.847033396 0.789946180 0.736706451
#>  [31] 0.687054902 0.640749701 0.597565316 0.557291414 0.519731838 0.484703651
#>  [37] 0.452036247 0.421570517 0.393158075 0.366660537 0.341948844 0.318902636
#>  [43] 0.297409666 0.277365250 0.258671760 0.241238149 0.224979506 0.209816640
#>  [49] 0.195675701 0.182487813 0.170188745 0.158718594 0.148021492 0.138045338
#>  [55] 0.128741545 0.120064795 0.111972830 0.104426236 0.097388257 0.090824615
#>  [61] 0.084703340 0.078994618 0.073670645 0.068705490 0.064074970 0.059756532
#>  [67] 0.055729141 0.051973184 0.048470365 0.045203625 0.042157052 0.039315808
#>  [73] 0.036666054 0.034194884 0.031890264 0.029740967 0.027736525 0.025867176
#>  [79] 0.024123815 0.022497951 0.020981664 0.019567570 0.018248781 0.017018875
#>  [85] 0.015871859 0.014802149 0.013804534 0.012874154 0.012006480 0.011197283
#>  [91] 0.010442624 0.009738826 0.009082461 0.008470334 0.007899462 0.007367065
#>  [97] 0.006870549 0.006407497 0.005975653 0.005572914
#> 
#> $fit
#> $beta
#>               5.5729     5.1973     4.8470     4.5204   4.2157   3.9316
#> (Intercept) 19.40476 21.6686989 23.7800542 25.7491112 27.58546 29.29805
#> am           0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.00000
#> carb         0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.00000
#> cyl          0.00000 -0.3547961 -0.6856801 -0.9942637 -1.28205 -1.55044
#> disp         0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.00000
#> drat         0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.00000
#> gear         0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.00000
#> hp           0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.00000
#> qsec         0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.00000
#> vs           0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.00000
#> wt           0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.00000
#>                3.6666    3.4195    3.1890    2.9741    2.7737    2.5867
#> (Intercept) 30.895209 32.384728 33.773859 35.069367 36.277562 37.404329
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.800742 -2.034174 -2.251873 -2.454901 -2.644245 -2.820828
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                2.4124    2.2498    2.0982    1.9568    1.8249    1.7019
#> (Intercept) 38.455155 39.435159 40.349115 41.201472 41.798921 41.798921
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.985509 -3.139092 -3.282324 -3.415902 -3.509532 -3.509532
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.5872    1.4802    1.3805    1.2874    1.2006    1.1197
#> (Intercept) 41.798921 41.798921 41.798921 41.798921 41.798921 41.798921
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -3.509532 -3.509532 -3.509532 -3.509532 -3.509532 -3.509532
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.0443    0.9739    0.9082    0.8470    0.7899    0.7367
#> (Intercept) 41.798921 41.798921 41.903078 41.902604 41.902291 41.902083
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -3.509532 -3.509532 -2.041657 -2.041433 -2.041285 -2.041187
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000 -2.784808 -2.785089 -2.785274 -2.785397
#>                0.6871    0.6407       0.5976      0.5573     0.5197     0.4847
#> (Intercept) 41.901946 41.901856 41.898540793 41.83157005 41.7689197 41.7109052
#> am           0.000000  0.000000  0.000000000  0.00000000  0.0000000  0.0000000
#> carb         0.000000  0.000000 -0.004687771 -0.08335294 -0.1567967 -0.2250858
#> cyl         -2.041122 -2.041079 -2.037157300 -1.96598235 -1.8994727 -1.8377438
#> disp         0.000000  0.000000  0.000000000  0.00000000  0.0000000  0.0000000
#> drat         0.000000  0.000000  0.000000000  0.00000000  0.0000000  0.0000000
#> gear         0.000000  0.000000  0.000000000  0.00000000  0.0000000  0.0000000
#> hp           0.000000  0.000000  0.000000000  0.00000000  0.0000000  0.0000000
#> qsec         0.000000  0.000000  0.000000000  0.00000000  0.0000000  0.0000000
#> vs           0.000000  0.000000  0.000000000  0.00000000  0.0000000  0.0000000
#> wt          -2.785479 -2.785532 -2.787452708 -2.82640459 -2.8628450 -2.8965864
#>                 0.4520     0.4216     0.3932        0.3667     0.3419
#> (Intercept) 41.6563942 41.6055466 41.5581256 36.8831309544 10.0458423
#> am           0.0000000  0.0000000  0.0000000  2.8371014838  3.9225816
#> carb        -0.2889738 -0.3485612 -0.4041327 -0.9442534009 -1.1018950
#> cyl         -1.7798819 -1.7259122 -1.6755797 -1.2439977116  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0005209707  2.6666933
#> gear         0.0000000  0.0000000  0.0000000  0.0000000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000000  0.5718525
#> vs           0.0000000  0.0000000  0.0000000  0.0000000000  0.0000000
#> wt          -2.9282927 -2.9578683 -2.9854508 -2.2246255055 -2.4017959
#>                 0.3189    0.2974    0.2774     0.2587     0.2412     0.2250
#> (Intercept) 10.0460130 10.046157 10.046275 10.0463692 10.0464439 10.0465023
#> am           3.9226155  3.922638  3.922653  3.9226629  3.9226697  3.9226744
#> carb        -1.1016788 -1.101516 -1.101394 -1.1013017 -1.1012324 -1.1011801
#> cyl          0.0000000  0.000000  0.000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.000000  0.000000  0.0000000  0.0000000  0.0000000
#> drat         2.6660164  2.665506  2.665122  2.6648324  2.6646140  2.6644494
#> gear         0.0000000  0.000000  0.000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.000000  0.000000  0.0000000  0.0000000  0.0000000
#> qsec         0.5719911  0.572095  0.572173  0.5722315  0.5722755  0.5723086
#> vs           0.0000000  0.000000  0.000000  0.0000000  0.0000000  0.0000000
#> wt          -2.4020854 -2.402305 -2.402470 -2.4025957 -2.4026904 -2.4027620
#>                 0.2098     0.1957     0.1825    0.1702    0.1587    0.1480
#> (Intercept) 10.0465478 10.0465828 10.0466098 10.088477 10.087283 10.086351
#> am           3.9226776  3.9226799  3.9226815  2.317549  2.317343  2.317183
#> carb        -1.1011408 -1.1011111 -1.1010887 -2.078173 -2.078316 -2.078428
#> cyl          0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> disp         0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> drat         2.6643253  2.6642317  2.6641611  2.885513  2.885237  2.885022
#> gear         0.0000000  0.0000000  0.0000000  2.533603  2.534162  2.534597
#> hp           0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> qsec         0.5723336  0.5723523  0.5723665  0.000000  0.000000  0.000000
#> vs           0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> wt          -2.4028160 -2.4028568 -2.4028877 -1.254492 -1.254299 -1.254148
#>                0.1380    0.1287    0.1201      0.1120      0.1044      0.0974
#> (Intercept) 10.085623 10.085055 10.084611 -47.8646252 -47.8705833 -47.8763971
#> am           2.317059  2.316962  2.316887   3.6112031   3.6112444   3.6112848
#> carb        -2.078515 -2.078583 -2.078635  -3.7888900  -3.7891645  -3.7894323
#> cyl          0.000000  0.000000  0.000000   3.5535193   3.5539007   3.5542728
#> disp         0.000000  0.000000  0.000000   0.0000000   0.0000000   0.0000000
#> drat         2.884855  2.884724  2.884623   5.3971869   5.3974755   5.3977571
#> gear         2.534936  2.535200  2.535406   6.8637511   6.8643942   6.8650218
#> hp           0.000000  0.000000  0.000000   0.0000000   0.0000000   0.0000000
#> qsec         0.000000  0.000000  0.000000   0.8787568   0.8787961   0.8788345
#> vs           0.000000  0.000000  0.000000   0.0000000   0.0000000   0.0000000
#> wt          -1.254030 -1.253938 -1.253866  -1.2835454  -1.2834508  -1.2833586
#>                 0.0908      0.0847      0.0790      0.0737       0.0687
#> (Intercept) -47.882070 -47.8876056 -47.8930070 -47.9029095 -66.27473238
#> am            3.611324   3.6113625   3.6113999   3.6115515   3.38606751
#> carb         -3.789694  -3.7899487  -3.7901975  -3.7906925  -3.90515823
#> cyl           3.554636   3.5549902   3.5553359   3.5559922   4.59216919
#> disp          0.000000   0.0000000   0.0000000   0.0000000   0.02111633
#> drat          5.398032   5.3983000   5.3985616   5.3989813   6.37253988
#> gear          6.865634   6.8662318   6.8668149   6.8679300   8.59283436
#> hp            0.000000   0.0000000   0.0000000   0.0000000  -0.01883021
#> qsec          0.878872   0.8789086   0.8789442   0.8790041   1.22251763
#> vs            0.000000   0.0000000   0.0000000   0.0000000   0.00000000
#> wt           -1.283269  -1.2831807  -1.2830950  -1.2829114  -2.96710855
#>                   0.0641       0.0598       0.0557       0.0520       0.0485
#> (Intercept) -66.28024593 -66.28562797 -66.29088158 -66.29600977 -66.30101547
#> am            3.38610140   3.38613454   3.38616693   3.38619859   3.38622955
#> carb         -3.90521378  -3.90526789  -3.90532059  -3.90537191  -3.90542189
#> cyl           4.59248636   4.59279593   4.59309807   4.59339297   4.59368080
#> disp          0.02112136   0.02112627   0.02113107   0.02113575   0.02114033
#> drat          6.37276135   6.37297750   6.37318843   6.37339427   6.37359514
#> gear          8.59328417   8.59372311   8.59415144   8.59456941   8.59497725
#> hp           -0.01883390  -0.01883750  -0.01884102  -0.01884445  -0.01884780
#> qsec          1.22263778   1.22275514   1.22286977   1.22298172   1.22309106
#> vs            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> wt           -2.96755975  -2.96800053  -2.96843114  -2.96885180  -2.96926274
#>                   0.0452       0.0422       0.0393       0.0367       0.0342
#> (Intercept) -66.30590155 -66.31067081 -66.31532599 -66.31986976 -66.32430474
#> am            3.38625980   3.38628938   3.38631829   3.38634655   3.38637417
#> carb         -3.90547056  -3.90551796  -3.90556411  -3.90560906  -3.90565282
#> cyl           4.59396171   4.59423588   4.59450346   4.59476461   4.59501947
#> disp          0.02114480   0.02114916   0.02115343   0.02115759   0.02116166
#> drat          6.37379116   6.37398245   6.37416912   6.37435127   6.37452901
#> gear          8.59537522   8.59576355   8.59614247   8.59651220   8.59687296
#> hp           -0.01885108  -0.01885428  -0.01885740  -0.01886045  -0.01886342
#> qsec          1.22319785   1.22330215   1.22340401   1.22350350   1.22360065
#> vs            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> wt           -2.96966418  -2.97005633  -2.97043940  -2.97081360  -2.97117913
#>                   0.0319       0.0297       0.0277       0.0259       0.0241
#> (Intercept) -66.32863348 -66.33285849 -66.33698218 -66.34100696 -66.34493514
#> am            3.38640117   3.38642756   3.38645335   3.38647856   3.38650320
#> carb         -3.90569544  -3.90573694  -3.90577735  -3.90581669  -3.90585500
#> cyl           4.59526821   4.59551095   4.59574785   4.59597904   4.59620466
#> disp          0.02116563   0.02116951   0.02117329   0.02117699   0.02118060
#> drat          6.37470246   6.37487170   6.37503684   6.37519798   6.37535522
#> gear          8.59722496   8.59756841   8.59790352   8.59823048   8.59854949
#> hp           -0.01886633  -0.01886917  -0.01887194  -0.01887464  -0.01887728
#> qsec          1.22369553   1.22378819   1.22387869   1.22396706   1.22405336
#> vs            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> wt           -2.97153619  -2.97188496  -2.97222564  -2.97255841  -2.97288346
#>                   0.0225       0.0210       0.0196       0.0182       0.0170
#> (Intercept) -66.34876899 -66.82379095 -66.82777232 -66.83149917 -66.83498580
#> am            3.38652728   3.58502136   3.58553224   3.58600549   3.58644407
#> carb         -3.90589231  -3.86348445  -3.86333905  -3.86319926  -3.86306466
#> cyl           4.59642483   4.67555148   4.67579291   4.67601668   4.67622398
#> disp          0.02118413   0.02137347   0.02137831   0.02138298   0.02138749
#> drat          6.37550864   6.38204387   6.38213993   6.38222775   6.38230782
#> gear          8.59886073   8.50117277   8.50100963   8.50085717   8.50071416
#> hp           -0.01887986  -0.01941497  -0.01941692  -0.01941883  -0.01942068
#> qsec          1.22413764   1.22604613   1.22621264   1.22637089   1.22652134
#> vs            0.00000000   0.32230876   0.32269223   0.32304596   0.32337246
#> wt           -2.97320094  -2.98516006  -2.98573262  -2.98628429  -2.98681579
#>                   0.0159       0.0148       0.0138       0.0129       0.0120
#> (Intercept) -66.83824581 -66.84129209 -66.84413684 -66.84679162 -66.84926734
#> am            3.58685069   3.58722789   3.58757797   3.58790309   3.58820522
#> carb         -3.86293488  -3.86280957  -3.86268842  -3.86257117  -3.86245756
#> cyl           4.67641590   4.67659345   4.67675758   4.67690919   4.67704910
#> disp          0.02139185   0.02139606   0.02140013   0.02140405   0.02140783
#> drat          6.38238061   6.38244654   6.38250603   6.38255948   6.38260726
#> gear          8.50057948   8.50045215   8.50033128   8.50021607   8.50010582
#> hp           -0.01942249  -0.01942425  -0.01942596  -0.01942762  -0.01942923
#> qsec          1.22666438   1.22680041   1.22692982   1.22705294   1.22717012
#> vs            0.32367401   0.32395271   0.32421048   0.32444908   0.32467010
#> wt           -2.98732784  -2.98782111  -2.98829629  -2.98875402  -2.98919494
#>                   0.0112       0.0104       0.0097       0.0091       0.0085
#> (Intercept) -66.85157431 -66.85372224 -66.85572029 -66.85757711 -66.85930081
#> am            3.58848616   3.58874760   3.58899108   3.58921800   3.58942967
#> carb         -3.86234736  -3.86224038  -3.86213641  -3.86203530  -3.86193690
#> cyl           4.67717807   4.67729682   4.67740603   4.67750631   4.67759825
#> disp          0.02141148   0.02141500   0.02141839   0.02142165   0.02142480
#> drat          6.38264973   6.38268721   6.38272002   6.38274846   6.38277281
#> gear          8.49999989   8.49989772   8.49979881   8.49970272   8.49960907
#> hp           -0.01943079  -0.01943231  -0.01943378  -0.01943520  -0.01943658
#> qsec          1.22728166   1.22738788   1.22748903   1.22758540   1.22767723
#> vs            0.32487502   0.32506517   0.32524177   0.32540596   0.32555874
#> wt           -2.98961965  -2.99002875  -2.99042280  -2.99080235  -2.99116794
#>                   0.0079       0.0074       0.0069       0.0064       0.0060
#> (Intercept) -66.86089905 -66.86237903 -66.86374752 -66.86501089 -66.86617513
#> am            3.58962728   3.58981194   3.58998463   3.59014630   3.59029780
#> carb         -3.86184106  -3.86174766  -3.86165658  -3.86156772  -3.86148098
#> cyl           4.67768237   4.67775919   4.67782918   4.67789276   4.67795035
#> disp          0.02142783   0.02143074   0.02143355   0.02143625   0.02143885
#> drat          6.38279332   6.38281024   6.38282381   6.38283424   6.38284175
#> gear          8.49951749   8.49942771   8.49933944   8.49925246   8.49916656
#> hp           -0.01943791  -0.01943919  -0.01944043  -0.01944163  -0.01944278
#> qsec          1.22776475   1.22784818   1.22792773   1.22800360   1.22807597
#> vs            0.32570105   0.32583374   0.32595759   0.32607331   0.32618154
#> wt           -2.99152007  -2.99185923  -2.99218590  -2.99250052  -2.99280354
#>                   0.0056
#> (Intercept) -66.86724587
#> am            3.59043988
#> carb         -3.86139627
#> cyl           4.67800233
#> disp          0.02144135
#> drat          6.38284651
#> gear          8.49908159
#> hp           -0.01944389
#> qsec          1.22814502
#> vs            0.32628289
#> wt           -2.99309537
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1   1   1   1  45   1   1   1   1   1   6  15  15  14
#>  [37]  14  14  14  46 117   1   1   1   1   1   1   1   1   1  66   1   1   1
#>  [55]   1   1 221   1   1   1   1   1   2 319   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1  52   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.572914137 5.197318376 4.847036512 4.520362474 4.215705172 3.931580753
#>   [7] 3.666605369 3.419488439 3.189026363 2.974096659 2.773652498 2.586717603
#>  [13] 2.412381494 2.249795055 2.098166398 1.956757005 1.824878132 1.701887454
#>  [19] 1.587185935 1.480214916 1.380453384 1.287415446 1.200647954 1.119728300
#>  [25] 1.044262361 0.973882573 0.908246147 0.847033396 0.789946180 0.736706451
#>  [31] 0.687054902 0.640749701 0.597565316 0.557291414 0.519731838 0.484703651
#>  [37] 0.452036247 0.421570517 0.393158075 0.366660537 0.341948844 0.318902636
#>  [43] 0.297409666 0.277365250 0.258671760 0.241238149 0.224979506 0.209816640
#>  [49] 0.195675701 0.182487813 0.170188745 0.158718594 0.148021492 0.138045338
#>  [55] 0.128741545 0.120064795 0.111972830 0.104426236 0.097388257 0.090824615
#>  [61] 0.084703340 0.078994618 0.073670645 0.068705490 0.064074970 0.059756532
#>  [67] 0.055729141 0.051973184 0.048470365 0.045203625 0.042157052 0.039315808
#>  [73] 0.036666054 0.034194884 0.031890264 0.029740967 0.027736525 0.025867176
#>  [79] 0.024123815 0.022497951 0.020981664 0.019567570 0.018248781 0.017018875
#>  [85] 0.015871859 0.014802149 0.013804534 0.012874154 0.012006480 0.011197283
#>  [91] 0.010442624 0.009738826 0.009082461 0.008470334 0.007899462 0.007367065
#>  [97] 0.006870549 0.006407497 0.005975653 0.005572914
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 26
#> 
#> $loss
#>   [1] 834.58952 709.38592 604.63449 517.39200 445.11742 385.61822 337.00385
#>   [8] 297.64526 266.13982 241.28088 222.03136 207.50078 196.92544 189.65115
#>  [15] 185.11831 182.84892 182.38471 182.38471 182.38471 182.38471 182.38471
#>  [22] 182.38471 182.38471 182.38471 182.38471 182.38471 123.87618 123.87618
#>  [29] 123.87618 123.87617 123.87617 123.87617 123.67101 120.49048 117.96796
#>  [36] 116.00966 114.51539 113.41603 112.64676  98.61428  84.59086  84.59086
#>  [43]  84.59085  84.59085  84.59085  84.59085  84.59085  84.59084  84.59084
#>  [50]  84.59084  79.76754  79.76754  79.76753  79.76753  79.76753  79.76753
#>  [57]  62.68205  62.68204  62.68202  62.68200  62.68199  62.68197  62.68195
#>  [64]  58.34096  58.34093  58.34091  58.34089  58.34087  58.34085  58.34083
#>  [71]  58.34082  58.34080  58.34079  58.34077  58.34076  58.34074  58.34073
#>  [78]  58.34072  58.34071  58.34070  58.26479  58.26477  58.26474  58.26472
#>  [85]  58.26471  58.26469  58.26467  58.26466  58.26465  58.26464  58.26462
#>  [92]  58.26461  58.26461  58.26460  58.26459  58.26458  58.26458  58.26457
#>  [99]  58.26456  58.26456
#> 
#> $linear.predictors
#>      5.5729   5.1973   4.8470   4.5204   4.2157   3.9316   3.6666   3.4195
#> 1  19.40476 19.53992 19.66597 19.78353 19.89316 19.99541 20.09076 20.17969
#> 2  19.40476 19.53992 19.66597 19.78353 19.89316 19.99541 20.09076 20.17969
#> 3  19.40476 20.24951 21.03733 21.77206 22.45726 23.09629 23.69224 24.24803
#> 4  19.40476 19.53992 19.66597 19.78353 19.89316 19.99541 20.09076 20.17969
#> 5  19.40476 19.53992 19.66597 19.78353 19.89316 19.99541 20.09076 20.17969
#> 6  19.40476 18.83033 18.29461 17.79500 17.32906 16.89453 16.48928 16.11134
#> 7  19.40476 20.24951 21.03733 21.77206 22.45726 23.09629 23.69224 24.24803
#> 8  19.40476 19.53992 19.66597 19.78353 19.89316 19.99541 20.09076 20.17969
#> 9  19.40476 19.53992 19.66597 19.78353 19.89316 19.99541 20.09076 20.17969
#> 10 19.40476 18.83033 18.29461 17.79500 17.32906 16.89453 16.48928 16.11134
#> 11 19.40476 18.83033 18.29461 17.79500 17.32906 16.89453 16.48928 16.11134
#> 12 19.40476 18.83033 18.29461 17.79500 17.32906 16.89453 16.48928 16.11134
#> 13 19.40476 18.83033 18.29461 17.79500 17.32906 16.89453 16.48928 16.11134
#> 14 19.40476 18.83033 18.29461 17.79500 17.32906 16.89453 16.48928 16.11134
#> 15 19.40476 18.83033 18.29461 17.79500 17.32906 16.89453 16.48928 16.11134
#> 16 19.40476 20.24951 21.03733 21.77206 22.45726 23.09629 23.69224 24.24803
#> 17 19.40476 20.24951 21.03733 21.77206 22.45726 23.09629 23.69224 24.24803
#> 18 19.40476 18.83033 18.29461 17.79500 17.32906 16.89453 16.48928 16.11134
#> 19 19.40476 20.24951 21.03733 21.77206 22.45726 23.09629 23.69224 24.24803
#> 20 19.40476 19.53992 19.66597 19.78353 19.89316 19.99541 20.09076 20.17969
#> 21 19.40476 18.83033 18.29461 17.79500 17.32906 16.89453 16.48928 16.11134
#>      3.1890   2.9741   2.7737   2.5867   2.4124   2.2498   2.0982   1.9568
#> 1  20.26262 20.33996 20.41209 20.47936 20.54210 20.60061 20.65517 20.70606
#> 2  20.26262 20.33996 20.41209 20.47936 20.54210 20.60061 20.65517 20.70606
#> 3  24.76637 25.24976 25.70058 26.12102 26.51312 26.87879 27.21982 27.53786
#> 4  20.26262 20.33996 20.41209 20.47936 20.54210 20.60061 20.65517 20.70606
#> 5  20.26262 20.33996 20.41209 20.47936 20.54210 20.60061 20.65517 20.70606
#> 6  15.75887 15.43016 15.12360 14.83771 14.57108 14.32242 14.09052 13.87425
#> 7  24.76637 25.24976 25.70058 26.12102 26.51312 26.87879 27.21982 27.53786
#> 8  20.26262 20.33996 20.41209 20.47936 20.54210 20.60061 20.65517 20.70606
#> 9  20.26262 20.33996 20.41209 20.47936 20.54210 20.60061 20.65517 20.70606
#> 10 15.75887 15.43016 15.12360 14.83771 14.57108 14.32242 14.09052 13.87425
#> 11 15.75887 15.43016 15.12360 14.83771 14.57108 14.32242 14.09052 13.87425
#> 12 15.75887 15.43016 15.12360 14.83771 14.57108 14.32242 14.09052 13.87425
#> 13 15.75887 15.43016 15.12360 14.83771 14.57108 14.32242 14.09052 13.87425
#> 14 15.75887 15.43016 15.12360 14.83771 14.57108 14.32242 14.09052 13.87425
#> 15 15.75887 15.43016 15.12360 14.83771 14.57108 14.32242 14.09052 13.87425
#> 16 24.76637 25.24976 25.70058 26.12102 26.51312 26.87879 27.21982 27.53786
#> 17 24.76637 25.24976 25.70058 26.12102 26.51312 26.87879 27.21982 27.53786
#> 18 15.75887 15.43016 15.12360 14.83771 14.57108 14.32242 14.09052 13.87425
#> 19 24.76637 25.24976 25.70058 26.12102 26.51312 26.87879 27.21982 27.53786
#> 20 20.26262 20.33996 20.41209 20.47936 20.54210 20.60061 20.65517 20.70606
#> 21 15.75887 15.43016 15.12360 14.83771 14.57108 14.32242 14.09052 13.87425
#>      1.8249   1.7019   1.5872   1.4802   1.3805   1.2874   1.2006   1.1197
#> 1  20.74173 20.74173 20.74173 20.74173 20.74173 20.74173 20.74173 20.74173
#> 2  20.74173 20.74173 20.74173 20.74173 20.74173 20.74173 20.74173 20.74173
#> 3  27.76079 27.76079 27.76079 27.76079 27.76079 27.76079 27.76079 27.76079
#> 4  20.74173 20.74173 20.74173 20.74173 20.74173 20.74173 20.74173 20.74173
#> 5  20.74173 20.74173 20.74173 20.74173 20.74173 20.74173 20.74173 20.74173
#> 6  13.72266 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266
#> 7  27.76079 27.76079 27.76079 27.76079 27.76079 27.76079 27.76079 27.76079
#> 8  20.74173 20.74173 20.74173 20.74173 20.74173 20.74173 20.74173 20.74173
#> 9  20.74173 20.74173 20.74173 20.74173 20.74173 20.74173 20.74173 20.74173
#> 10 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266
#> 11 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266
#> 12 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266
#> 13 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266
#> 14 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266
#> 15 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266
#> 16 27.76079 27.76079 27.76079 27.76079 27.76079 27.76079 27.76079 27.76079
#> 17 27.76079 27.76079 27.76079 27.76079 27.76079 27.76079 27.76079 27.76079
#> 18 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266
#> 19 27.76079 27.76079 27.76079 27.76079 27.76079 27.76079 27.76079 27.76079
#> 20 20.74173 20.74173 20.74173 20.74173 20.74173 20.74173 20.74173 20.74173
#> 21 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266 13.72266
#>      1.0443   0.9739   0.9082   0.8470   0.7899   0.7367   0.6871   0.6407
#> 1  20.74173 20.74173 22.35694 22.35707 22.35716 22.35722 22.35726 22.35729
#> 2  20.74173 20.74173 21.64681 21.64688 21.64692 21.64695 21.64696 21.64698
#> 3  27.76079 27.76079 27.27570 27.27547 27.27531 27.27521 27.27515 27.27510
#> 4  20.74173 20.74173 20.69998 20.69995 20.69992 20.69991 20.69990 20.69989
#> 5  20.74173 20.74173 20.01770 20.01760 20.01753 20.01749 20.01746 20.01744
#> 6  13.72266 13.72266 15.62806 15.62837 15.62858 15.62872 15.62881 15.62887
#> 7  27.76079 27.76079 24.96431 24.96384 24.96354 24.96333 24.96320 24.96311
#> 8  20.74173 20.74173 20.07340 20.07330 20.07324 20.07320 20.07317 20.07315
#> 9  20.74173 20.74173 20.07340 20.07330 20.07324 20.07320 20.07317 20.07315
#> 10 13.72266 13.72266 14.23566 14.23583 14.23595 14.23602 14.23607 14.23611
#> 11 13.72266 13.72266 15.18249 15.18276 15.18294 15.18306 15.18314 15.18319
#> 12 13.72266 13.72266 15.04325 15.04351 15.04368 15.04379 15.04386 15.04391
#> 13 13.72266 13.72266 10.94958 10.94943 10.94932 10.94925 10.94921 10.94918
#> 14 13.72266 13.72266 10.46503 10.46482 10.46468 10.46459 10.46453 10.46450
#> 15 13.72266 13.72266 10.68503 10.68484 10.68472 10.68464 10.68459 10.68455
#> 16 27.76079 27.76079 27.60987 27.60968 27.60955 27.60946 27.60941 27.60937
#> 17 27.76079 27.76079 28.62633 28.62623 28.62617 28.62613 28.62611 28.62609
#> 18 13.72266 13.72266 14.87616 14.87640 14.87656 14.87666 14.87673 14.87678
#> 19 27.76079 27.76079 29.52304 29.52303 29.52303 29.52303 29.52303 29.52303
#> 20 20.74173 20.74173 21.93922 21.93931 21.93937 21.93941 21.93944 21.93946
#> 21 13.72266 13.72266 15.62806 15.62837 15.62858 15.62872 15.62881 15.62887
#>      0.5976   0.5573   0.5197   0.4847   0.4520   0.4216   0.3932   0.3667
#> 1  22.35372 22.29708 22.24424 22.19504 22.14908 22.10621 22.06624 22.65275
#> 2  21.64292 21.57635 21.51422 21.45641 21.40237 21.35196 21.30495 22.08547
#> 3  27.27833 27.32703 27.37243 27.41476 27.45425 27.49108 27.52543 28.64086
#> 4  20.70925 20.86543 21.01124 21.14683 21.27367 21.39197 21.50229 21.32432
#> 5  20.02632 20.17296 20.30984 20.43717 20.55624 20.66729 20.77086 20.77912
#> 6  15.63133 15.68004 15.72559 15.76780 15.80744 15.84441 15.87890 15.21389
#> 7  24.96006 24.89776 24.83947 24.78551 24.73480 24.68749 24.64337 23.01311
#> 8  20.06801 19.97943 19.89671 19.81984 19.74788 19.68076 19.61817 17.99146
#> 9  20.06801 19.97943 19.89671 19.81984 19.74788 19.68076 19.61817 17.99146
#> 10 14.24229 14.35019 14.45097 14.54459 14.63227 14.71404 14.79031 15.04576
#> 11 15.19002 15.31116 15.42434 15.52943 15.62789 15.71972 15.80536 15.80214
#> 12 15.05065 15.16984 15.28119 15.38460 15.48147 15.57182 15.65609 15.69090
#> 13 10.94840 10.93168 10.91602 10.90153 10.88791 10.87520 10.86334 11.47638
#> 14 10.46339 10.43988 10.41788 10.39753 10.37838 10.36053 10.34387 11.08933
#> 15 10.68360 10.66317 10.64405 10.62636 10.60972 10.59420 10.57972 11.26520
#> 16 27.61283 27.66620 27.71597 27.76235 27.80565 27.84603 27.88368 28.90794
#> 17 28.63025 28.69784 28.76091 28.81961 28.87448 28.92565 28.97337 29.72000
#> 18 14.87871 14.91691 14.95263 14.98572 15.01680 15.04579 15.07283 14.61352
#> 19 29.52312 29.52458 29.52595 29.52722 29.52841 29.52952 29.53055 29.49184
#> 20 21.92623 21.70642 21.50122 21.31038 21.13189 20.96541 20.81015 20.43040
#> 21 15.61257 15.34662 15.09841 14.86746 14.65154 14.45017 14.26237 14.27415
#>      0.3419   0.3189   0.2974   0.2774   0.2587   0.2412   0.2250   0.2098
#> 1  23.08094 23.08089 23.08085 23.08082 23.08080 23.08079 23.08077 23.08076
#> 2  22.78871 22.78867 22.78864 22.78861 22.78859 22.78857 22.78856 22.78855
#> 3  28.20331 28.20303 28.20282 28.20266 28.20254 28.20245 28.20238 28.20233
#> 4  20.55240 20.55247 20.55252 20.55256 20.55259 20.55261 20.55263 20.55264
#> 5  19.55667 19.55698 19.55723 19.55741 19.55754 19.55765 19.55773 19.55779
#> 6  14.68208 14.68210 14.68213 14.68214 14.68215 14.68217 14.68217 14.68218
#> 7  23.82526 23.82547 23.82563 23.82574 23.82583 23.82590 23.82595 23.82599
#> 8  18.29442 18.29435 18.29429 18.29424 18.29421 18.29419 18.29417 18.29416
#> 9  18.63754 18.63754 18.63755 18.63755 18.63755 18.63755 18.63756 18.63756
#> 10 15.10183 15.10180 15.10179 15.10177 15.10176 15.10176 15.10175 15.10175
#> 11 16.03281 16.03291 16.03299 16.03305 16.03309 16.03313 16.03315 16.03317
#> 12 16.14146 16.14160 16.14171 16.14179 16.14185 16.14190 16.14194 16.14196
#> 13 11.12415 11.12418 11.12419 11.12421 11.12422 11.12422 11.12423 11.12423
#> 14 10.80141 10.80132 10.80124 10.80119 10.80115 10.80111 10.80109 10.80107
#> 15 11.37575 11.37547 11.37525 11.37509 11.37497 11.37488 11.37481 11.37475
#> 16 29.59666 29.59638 29.59616 29.59600 29.59588 29.59579 29.59572 29.59567
#> 17 31.09254 31.09234 31.09218 31.09206 31.09197 31.09190 31.09185 31.09181
#> 18 15.17438 15.17391 15.17357 15.17330 15.17311 15.17296 15.17285 15.17276
#> 19 27.84846 27.84845 27.84844 27.84844 27.84843 27.84843 27.84843 27.84843
#> 20 19.22122 19.22162 19.22192 19.22214 19.22231 19.22244 19.22254 19.22261
#> 21 14.36799 14.36852 14.36892 14.36921 14.36944 14.36961 14.36973 14.36983
#>      0.1957   0.1825   0.1702   0.1587   0.1480   0.1380   0.1287   0.1201
#> 1  23.08076 23.08075 22.19448 22.19417 22.19393 22.19374 22.19360 22.19348
#> 2  22.78855 22.78854 21.87458 21.87432 21.87412 21.87396 21.87384 21.87375
#> 3  28.20229 28.20226 28.66107 28.66114 28.66121 28.66125 28.66129 28.66132
#> 4  20.55265 20.55266 20.46530 20.46541 20.46550 20.46556 20.46561 20.46565
#> 5  19.55783 19.55786 19.23459 19.23483 19.23502 19.23517 19.23529 19.23538
#> 6  14.68218 14.68219 14.16055 14.16027 14.16004 14.15987 14.15973 14.15963
#> 7  23.82602 23.82604 23.42610 23.42638 23.42660 23.42677 23.42691 23.42701
#> 8  18.29415 18.29414 18.90595 18.90601 18.90604 18.90608 18.90610 18.90612
#> 9  18.63756 18.63756 18.90595 18.90601 18.90604 18.90608 18.90610 18.90612
#> 10 15.10174 15.10174 15.20751 15.20750 15.20749 15.20749 15.20749 15.20748
#> 11 16.03319 16.03320 15.63404 15.63396 15.63390 15.63386 15.63382 15.63380
#> 12 16.14198 16.14200 15.57131 15.57125 15.57120 15.57116 15.57113 15.57110
#> 13 11.12423 11.12424 11.24506 11.24518 11.24527 11.24534 11.24539 11.24544
#> 14 10.80106 10.80105 11.22877 11.22890 11.22900 11.22908 11.22914 11.22919
#> 15 11.37471 11.37468 11.99154 11.99159 11.99163 11.99166 11.99169 11.99171
#> 16 29.59563 29.59560 29.47527 29.47527 29.47526 29.47525 29.47525 29.47525
#> 17 31.09178 31.09176 30.33714 30.33702 30.33692 30.33685 30.33680 30.33675
#> 18 15.17270 15.17265 15.32231 15.32193 15.32163 15.32141 15.32123 15.32109
#> 19 27.84843 27.84843 29.89803 29.89839 29.89867 29.89889 29.89906 29.89919
#> 20 19.22266 19.22270 19.57562 19.57569 19.57574 19.57578 19.57582 19.57584
#> 21 14.36990 14.36995 14.18484 14.18480 14.18477 14.18474 14.18472 14.18471
#>      0.1120   0.1044   0.0974   0.0908   0.0847   0.0790   0.0737   0.0687
#> 1  21.51762 21.51748 21.51735 21.51723 21.51710 21.51698 21.51675 21.92407
#> 2  21.68241 21.68233 21.68224 21.68216 21.68208 21.68200 21.68185 21.85206
#> 3  27.78178 27.78175 27.78172 27.78169 27.78166 27.78163 27.78163 26.87719
#> 4  19.83862 19.83857 19.83851 19.83846 19.83840 19.83835 19.83828 20.38223
#> 5  18.48249 18.48239 18.48230 18.48220 18.48212 18.48203 18.48192 17.96695
#> 6  12.66144 12.66125 12.66107 12.66089 12.66072 12.66054 12.66021 12.83692
#> 7  23.46401 23.46393 23.46386 23.46378 23.46371 23.46364 23.46342 23.46889
#> 8  18.57876 18.57874 18.57872 18.57871 18.57869 18.57867 18.57856 18.39754
#> 9  19.10602 19.10602 19.10602 19.10603 19.10603 19.10604 19.10596 19.13106
#> 10 16.42381 16.42397 16.42412 16.42427 16.42441 16.42455 16.42484 15.71947
#> 11 17.03597 17.03610 17.03623 17.03635 17.03647 17.03659 17.03683 16.97279
#> 12 17.32330 17.32345 17.32359 17.32374 17.32388 17.32401 17.32428 17.31344
#> 13 10.87441 10.87439 10.87436 10.87434 10.87431 10.87429 10.87427 11.80230
#> 14 10.88828 10.88828 10.88829 10.88829 10.88829 10.88830 10.88833 11.09480
#> 15 11.87953 11.87958 11.87962 11.87967 11.87971 11.87975 11.87985 11.60110
#> 16 29.93289 29.93295 29.93300 29.93306 29.93311 29.93317 29.93329 29.63999
#> 17 31.53485 31.53493 31.53501 31.53509 31.53517 31.53524 31.53538 31.99917
#> 18 14.74356 14.74353 14.74350 14.74347 14.74344 14.74342 14.74332 14.61268
#> 19 29.95801 29.95818 29.95835 29.95851 29.95867 29.95883 29.95916 30.71000
#> 20 18.25624 18.25609 18.25595 18.25582 18.25568 18.25555 18.25530 17.76288
#> 21 15.53600 15.53609 15.53617 15.53626 15.53634 15.53642 15.53655 15.43446
#>      0.0641   0.0598   0.0557   0.0520   0.0485   0.0452   0.0422   0.0393
#> 1  21.92412 21.92418 21.92424 21.92429 21.92435 21.92440 21.92445 21.92450
#> 2  21.85207 21.85208 21.85210 21.85211 21.85212 21.85212 21.85213 21.85214
#> 3  26.87696 26.87674 26.87653 26.87632 26.87611 26.87591 26.87572 26.87552
#> 4  20.38237 20.38251 20.38265 20.38278 20.38291 20.38304 20.38316 20.38328
#> 5  17.96686 17.96677 17.96668 17.96659 17.96651 17.96643 17.96635 17.96627
#> 6  12.83699 12.83705 12.83710 12.83716 12.83722 12.83727 12.83733 12.83738
#> 7  23.46889 23.46889 23.46889 23.46889 23.46889 23.46889 23.46889 23.46889
#> 8  18.39741 18.39729 18.39716 18.39704 18.39693 18.39681 18.39670 18.39659
#> 9  19.13100 19.13094 19.13089 19.13083 19.13078 19.13073 19.13068 19.13063
#> 10 15.71933 15.71920 15.71907 15.71894 15.71882 15.71870 15.71858 15.71846
#> 11 16.97283 16.97287 16.97291 16.97295 16.97299 16.97302 16.97306 16.97309
#> 12 17.31351 17.31357 17.31364 17.31370 17.31376 17.31382 17.31387 17.31393
#> 13 11.80251 11.80271 11.80291 11.80310 11.80329 11.80348 11.80366 11.80384
#> 14 11.09483 11.09485 11.09488 11.09491 11.09494 11.09496 11.09499 11.09501
#> 15 11.60101 11.60092 11.60084 11.60076 11.60067 11.60060 11.60052 11.60044
#> 16 29.63993 29.63987 29.63981 29.63975 29.63969 29.63963 29.63957 29.63952
#> 17 31.99932 31.99947 31.99961 31.99975 31.99988 32.00001 32.00014 32.00027
#> 18 14.61263 14.61259 14.61254 14.61250 14.61245 14.61241 14.61237 14.61233
#> 19 30.71018 30.71034 30.71051 30.71067 30.71082 30.71097 30.71112 30.71127
#> 20 17.76272 17.76256 17.76240 17.76225 17.76210 17.76196 17.76182 17.76168
#> 21 15.43453 15.43460 15.43466 15.43472 15.43478 15.43485 15.43490 15.43496
#>      0.0367   0.0342   0.0319   0.0297   0.0277   0.0259   0.0241   0.0225
#> 1  21.92455 21.92460 21.92464 21.92469 21.92473 21.92478 21.92482 21.92486
#> 2  21.85215 21.85216 21.85217 21.85218 21.85219 21.85220 21.85220 21.85221
#> 3  26.87534 26.87516 26.87498 26.87481 26.87464 26.87447 26.87431 26.87415
#> 4  20.38340 20.38351 20.38363 20.38374 20.38384 20.38395 20.38405 20.38415
#> 5  17.96619 17.96612 17.96604 17.96597 17.96590 17.96584 17.96577 17.96571
#> 6  12.83743 12.83748 12.83753 12.83758 12.83762 12.83767 12.83772 12.83776
#> 7  23.46889 23.46889 23.46889 23.46889 23.46889 23.46889 23.46889 23.46889
#> 8  18.39648 18.39637 18.39627 18.39617 18.39607 18.39598 18.39588 18.39579
#> 9  19.13058 19.13054 19.13049 19.13044 19.13040 19.13036 19.13032 19.13028
#> 10 15.71835 15.71824 15.71813 15.71802 15.71792 15.71782 15.71772 15.71762
#> 11 16.97312 16.97316 16.97319 16.97322 16.97325 16.97328 16.97331 16.97334
#> 12 17.31399 17.31404 17.31409 17.31414 17.31419 17.31424 17.31429 17.31433
#> 13 11.80401 11.80418 11.80434 11.80450 11.80466 11.80482 11.80497 11.80511
#> 14 11.09504 11.09506 11.09508 11.09511 11.09513 11.09515 11.09517 11.09519
#> 15 11.60037 11.60030 11.60023 11.60016 11.60010 11.60003 11.59997 11.59990
#> 16 29.63946 29.63941 29.63936 29.63931 29.63926 29.63921 29.63917 29.63912
#> 17 32.00039 32.00051 32.00063 32.00074 32.00085 32.00096 32.00107 32.00117
#> 18 14.61229 14.61225 14.61221 14.61218 14.61214 14.61211 14.61207 14.61204
#> 19 30.71141 30.71154 30.71168 30.71181 30.71194 30.71206 30.71219 30.71230
#> 20 17.76155 17.76141 17.76129 17.76116 17.76104 17.76092 17.76080 17.76069
#> 21 15.43502 15.43507 15.43513 15.43518 15.43523 15.43528 15.43533 15.43538
#>      0.0210   0.0196   0.0182   0.0170   0.0159   0.0148   0.0138   0.0129
#> 1  21.89897 21.89905 21.89913 21.89921 21.89928 21.89935 21.89943 21.89949
#> 2  21.82434 21.82437 21.82440 21.82442 21.82445 21.82447 21.82449 21.82451
#> 3  26.89171 26.89156 26.89142 26.89128 26.89114 26.89101 26.89088 26.89076
#> 4  20.46431 20.46454 20.46476 20.46498 20.46518 20.46538 20.46556 20.46574
#> 5  18.03876 18.03880 18.03884 18.03887 18.03890 18.03893 18.03896 18.03899
#> 6  12.81790 12.81810 12.81830 12.81849 12.81868 12.81887 12.81905 12.81922
#> 7  23.33422 23.33411 23.33401 23.33392 23.33384 23.33377 23.33371 23.33365
#> 8  18.48204 18.48184 18.48165 18.48147 18.48129 18.48112 18.48096 18.48080
#> 9  19.21767 19.21757 19.21747 19.21738 19.21729 19.21720 19.21712 19.21703
#> 10 15.67027 15.67001 15.66976 15.66952 15.66928 15.66906 15.66884 15.66863
#> 11 16.93044 16.93040 16.93037 16.93034 16.93031 16.93028 16.93025 16.93022
#> 12 17.27160 17.27160 17.27160 17.27161 17.27161 17.27161 17.27161 17.27160
#> 13 11.81002 11.81021 11.81040 11.81058 11.81075 11.81092 11.81109 11.81125
#> 14 11.09055 11.09054 11.09054 11.09053 11.09053 11.09053 11.09053 11.09053
#> 15 11.58513 11.58500 11.58488 11.58475 11.58464 11.58453 11.58442 11.58431
#> 16 29.67016 29.67016 29.67015 29.67014 29.67013 29.67011 29.67009 29.67007
#> 17 32.03741 32.03766 32.03790 32.03813 32.03835 32.03855 32.03874 32.03891
#> 18 14.58963 14.58961 14.58959 14.58957 14.58956 14.58954 14.58953 14.58951
#> 19 30.66730 30.66721 30.66712 30.66704 30.66697 30.66691 30.66685 30.66680
#> 20 17.67885 17.67859 17.67834 17.67811 17.67789 17.67768 17.67749 17.67731
#> 21 15.52872 15.52905 15.52937 15.52966 15.52993 15.53018 15.53042 15.53064
#>      0.0120   0.0112   0.0104   0.0097   0.0091   0.0085   0.0079   0.0074
#> 1  21.89956 21.89963 21.89969 21.89975 21.89981 21.89987 21.89993 21.89999
#> 2  21.82453 21.82455 21.82457 21.82459 21.82461 21.82462 21.82464 21.82466
#> 3  26.89064 26.89052 26.89041 26.89030 26.89020 26.89010 26.89000 26.88990
#> 4  20.46592 20.46609 20.46625 20.46640 20.46655 20.46669 20.46683 20.46696
#> 5  18.03901 18.03904 18.03906 18.03908 18.03910 18.03912 18.03914 18.03916
#> 6  12.81939 12.81955 12.81971 12.81987 12.82002 12.82017 12.82031 12.82045
#> 7  23.33361 23.33356 23.33353 23.33350 23.33347 23.33345 23.33343 23.33342
#> 8  18.48065 18.48051 18.48037 18.48023 18.48010 18.47998 18.47986 18.47974
#> 9  19.21695 19.21687 19.21680 19.21672 19.21665 19.21658 19.21652 19.21645
#> 10 15.66843 15.66823 15.66805 15.66786 15.66769 15.66752 15.66736 15.66720
#> 11 16.93019 16.93016 16.93013 16.93011 16.93008 16.93005 16.93003 16.93000
#> 12 17.27160 17.27159 17.27159 17.27158 17.27157 17.27156 17.27156 17.27155
#> 13 11.81140 11.81155 11.81170 11.81183 11.81197 11.81210 11.81223 11.81235
#> 14 11.09053 11.09053 11.09053 11.09053 11.09054 11.09054 11.09054 11.09054
#> 15 11.58421 11.58412 11.58402 11.58393 11.58385 11.58376 11.58368 11.58360
#> 16 29.67005 29.67002 29.67000 29.66997 29.66994 29.66992 29.66989 29.66986
#> 17 32.03908 32.03924 32.03939 32.03953 32.03966 32.03979 32.03991 32.04002
#> 18 14.58950 14.58949 14.58948 14.58947 14.58946 14.58945 14.58944 14.58943
#> 19 30.66675 30.66671 30.66667 30.66664 30.66661 30.66658 30.66655 30.66653
#> 20 17.67714 17.67698 17.67683 17.67669 17.67656 17.67644 17.67632 17.67621
#> 21 15.53085 15.53105 15.53123 15.53140 15.53156 15.53171 15.53185 15.53198
#>      0.0069   0.0064   0.0060   0.0056
#> 1  21.90004 21.90009 21.90015 21.90020
#> 2  21.82467 21.82469 21.82470 21.82472
#> 3  26.88981 26.88972 26.88964 26.88956
#> 4  20.46709 20.46721 20.46733 20.46745
#> 5  18.03918 18.03920 18.03922 18.03923
#> 6  12.82059 12.82072 12.82085 12.82097
#> 7  23.33340 23.33339 23.33339 23.33338
#> 8  18.47963 18.47953 18.47943 18.47933
#> 9  19.21639 19.21633 19.21627 19.21622
#> 10 15.66705 15.66690 15.66676 15.66662
#> 11 16.92997 16.92995 16.92992 16.92990
#> 12 17.27154 17.27153 17.27151 17.27150
#> 13 11.81246 11.81258 11.81269 11.81279
#> 14 11.09055 11.09055 11.09055 11.09055
#> 15 11.58353 11.58346 11.58339 11.58332
#> 16 29.66983 29.66980 29.66977 29.66974
#> 17 32.04013 32.04023 32.04032 32.04041
#> 18 14.58942 14.58941 14.58941 14.58940
#> 19 30.66650 30.66648 30.66646 30.66644
#> 20 17.67611 17.67601 17.67592 17.67583
#> 21 15.53211 15.53222 15.53233 15.53244
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 22.8 21.4 18.1 14.3 22.8 19.2 17.8 16.4 17.3 15.2 10.4 10.4 14.7
#> [16] 32.4 33.9 13.3 30.4 19.7 15.0
#> 
#> $X
#>               am       carb        cyl       disp        drat       gear
#>  [1,]  1.2747549  0.4370867 -0.2399041 -0.6359244  0.88766350  0.4677072
#>  [2,]  1.2747549  0.4370867 -0.2399041 -0.6359244  0.88766350  0.4677072
#>  [3,]  1.2747549 -1.2839422 -1.4994004 -1.0626616  0.77273789  0.4677072
#>  [4,] -0.7844645 -1.2839422 -0.2399041  0.1683111 -0.99711646 -0.9354143
#>  [5,] -0.7844645 -1.2839422 -0.2399041 -0.1025029 -1.73264035 -0.9354143
#>  [6,] -0.7844645  0.4370867  1.0195922  1.0053725 -0.69830988 -0.9354143
#>  [7,] -0.7844645 -0.7102659 -1.4994004 -0.7934889  0.93363374  0.4677072
#>  [8,] -0.7844645  0.4370867 -0.2399041 -0.5735551  0.93363374  0.4677072
#>  [9,] -0.7844645  0.4370867 -0.2399041 -0.5735551  0.93363374  0.4677072
#> [10,] -0.7844645 -0.1365896  1.0195922  0.3143865 -1.02010158 -0.9354143
#> [11,] -0.7844645 -0.1365896  1.0195922  0.3143865 -1.02010158 -0.9354143
#> [12,] -0.7844645 -0.1365896  1.0195922  0.3143865 -1.02010158 -0.9354143
#> [13,] -0.7844645  0.4370867  1.0195922  1.9244988 -1.34189328 -0.9354143
#> [14,] -0.7844645  0.4370867  1.0195922  1.8260209 -1.18099743 -0.9354143
#> [15,] -0.7844645  0.4370867  1.0195922  1.6618913 -0.65233964 -0.9354143
#> [16,]  1.2747549 -1.2839422 -1.4994004 -1.3031116  1.30139569  0.4677072
#> [17,]  1.2747549 -1.2839422 -1.4994004 -1.3654809  1.62318739  0.4677072
#> [18,] -0.7844645  0.4370867  1.0195922  0.9233077  0.49691644 -0.9354143
#> [19,]  1.2747549 -0.7102659 -1.4994004 -1.1685252  0.58885692  1.8708287
#> [20,]  1.2747549  1.5844393 -0.2399041 -0.7590217  0.24408010  1.8708287
#> [21,]  1.2747549  2.7317918  1.0195922  0.5211899  0.06019913  1.8708287
#>               hp        qsec         vs          wt
#>  [1,] -0.6909863 -0.77006790 -0.8660254 -0.75844835
#>  [2,] -0.6909863 -0.46850285 -0.8660254 -0.51075115
#>  [3,] -0.9394074  0.38772649  1.1547005 -1.04985681
#>  [4,] -0.6909863  0.83468898  1.1547005 -0.18048823
#>  [5,] -0.7640513  1.25472601  1.1547005  0.05749535
#>  [6,]  1.2817692 -1.10394349 -0.8660254  0.16434512
#>  [7,] -0.9101814  2.69793018  1.1547005 -0.24362673
#>  [8,] -0.5010173  0.22078870  1.1547005  0.03806812
#>  [9,] -0.5010173  0.54389411  1.1547005  0.03806812
#> [10,]  0.3319239 -0.26386942 -0.8660254  0.65002590
#> [11,]  0.3319239 -0.15616762 -0.8660254  0.31976297
#> [12,]  0.3319239  0.05923599 -0.8660254  0.36833105
#> [13,]  0.6972490  0.04846581 -0.8660254  1.79623252
#> [14,]  0.8433791 -0.03769563 -0.8660254  1.96524943
#> [15,]  1.0625741 -0.25309924 -0.8660254  1.88851187
#> [16,] -1.3339585  0.85084425  1.1547005 -1.16642020
#> [17,] -1.3485715  1.08240313  1.1547005 -1.52096716
#> [18,]  1.2817692 -1.33550237 -0.8660254  0.42661274
#> [19,] -0.6471473 -0.53312393  1.1547005 -1.83374558
#> [20,]  0.2588589 -1.28703655 -0.8660254 -0.61274412
#> [21,]  2.5969395 -1.77169467 -0.8660254  0.16434512
#> attr(,"center")
#>  [1]   0.3809524   3.2380952   6.3809524 237.4904762   3.5138095   3.6666667
#>  [7] 157.2857143  17.8900000   0.4285714   3.4008095
#> attr(,"scale")
#>  [1]   0.4856209   1.7431434   1.5879364 121.8548586   0.4350640   0.7126966
#>  [7]  68.4321999   1.8569791   0.4948717   1.0294828
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 48
#> 
#> $lambda.min
#> [1] 0.2098166
#> 
#> $null.dev
#> [1] 39.74236
#> 
#> $Bias
#> [1] 5.773226
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 9.554401