Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 47.135692 42.952690 37.728403 33.094120 29.343369 26.342204 23.975396
#>   [8] 22.143913 20.762725 19.758910 19.070007 18.642585 18.431011 18.396375
#>  [15] 18.505563 18.730453 18.625878 18.306091 17.903746 17.469589 17.096368
#>  [22] 16.775867 16.554661 16.417513 16.265837 15.395719 14.202289 11.944597
#>  [29] 12.476804 13.558891 13.615134 13.606829 13.608967 13.642165 13.678339
#>  [36] 13.713900 13.705406 13.664213 14.031378 14.153165 13.943391 14.612118
#>  [43] 16.224186 15.359801 14.764193 15.434888 14.743536 14.742996 14.550706
#>  [50] 13.745843 14.679403 13.956843 14.855464 12.691798 12.693023 12.694208
#>  [57] 12.695355 12.696465 12.699171 11.177351 11.177972 11.178579 11.179171
#>  [64] 11.179749 11.497703 11.498348 11.498978 11.499592 11.500191 11.364035
#>  [71] 11.364394  9.333393  9.449304  9.595188  9.595996  9.927029 10.766285
#>  [78] 11.676351 11.677710 11.679040 11.680343 11.681618 11.682867 11.684090
#>  [85] 11.685288 11.686460 11.687609 11.747529 11.756716 11.757739 11.758739
#>  [92] 11.759717 11.760673 11.761608 11.762523 11.763417 11.764291 11.765146
#>  [99] 11.766090 11.776056
#> 
#> $cvse
#>   [1] 13.850563 13.519482 12.333602 11.059733  9.913680  8.879280  7.944349
#>   [8]  7.100156  6.341038  5.664114  5.069066  4.557856  4.134169  3.802325
#>  [15]  3.565467  3.423291  3.277158  3.192211  3.193848  3.181007  3.173843
#>  [22]  3.169758  3.169072  3.172526  3.132158  3.083410  3.110091  3.186567
#>  [29]  3.354785  3.834958  3.885979  3.897735  3.887467  3.879225  3.877969
#>  [36]  3.877564  3.867745  3.860552  4.009826  3.875114  3.818511  3.723302
#>  [43]  4.494987  4.471772  4.432833  4.360966  4.442870  4.442467  4.370193
#>  [50]  4.392408  4.396500  3.860943  3.953807  4.006509  4.006803  4.007083
#>  [57]  4.007351  4.007606  4.008239  3.477098  3.477192  3.477284  3.477373
#>  [64]  3.477461  3.564094  3.564175  3.564255  3.564333  3.564409  3.517684
#>  [71]  3.517728  2.606998  2.686890  2.703859  2.703954  2.684487  2.678804
#>  [78]  2.981586  2.981819  2.982045  2.982266  2.982480  2.982689  2.982891
#>  [85]  2.983088  2.983280  2.983466  2.976472  2.977670  2.977851  2.978027
#>  [92]  2.978198  2.978364  2.978526  2.978682  2.978835  2.978983  2.979126
#>  [99]  2.979274  2.979522
#> 
#> $fold
#>  [1]  8  9  7  2  6  4  1 10  6 10  2  5  3  9  4  5  1  7  1  3  8
#> 
#> $lambda
#>   [1] 5.726035563 5.340119930 4.980213720 4.644563983 4.331535916 4.039604892
#>   [7] 3.767349042 3.513442326 3.276648072 3.055812959 2.849861393 2.657790273
#>  [13] 2.478664104 2.311610439 2.155815632 2.010520874 1.875018496 1.748648525
#>  [19] 1.630795467 1.520885310 1.418382731 1.322788482 1.233636965 1.150493963
#>  [25] 1.072954520 1.000640977 0.933201124 0.870306491 0.811650747 0.756948203
#>  [31] 0.705932427 0.658354944 0.613984025 0.572603556 0.534011993 0.498021372
#>  [37] 0.464456398 0.433153592 0.403960489 0.376734904 0.351344233 0.327664807
#>  [43] 0.305581296 0.284986139 0.265779027 0.247866410 0.231161044 0.215581563
#>  [49] 0.201052087 0.187501850 0.174864852 0.163079547 0.152088531 0.141838273
#>  [55] 0.132278848 0.123363697 0.115049396 0.107295452 0.100064098 0.093320112
#>  [61] 0.087030649 0.081165075 0.075694820 0.070593243 0.065835494 0.061398402
#>  [67] 0.057260356 0.053401199 0.049802137 0.046445640 0.043315359 0.040396049
#>  [73] 0.037673490 0.035134423 0.032766481 0.030558130 0.028498614 0.026577903
#>  [79] 0.024786641 0.023116104 0.021558156 0.020105209 0.018750185 0.017486485
#>  [85] 0.016307955 0.015208853 0.014183827 0.013227885 0.012336370 0.011504940
#>  [91] 0.010729545 0.010006410 0.009332011 0.008703065 0.008116507 0.007569482
#>  [97] 0.007059324 0.006583549 0.006139840 0.005726036
#> 
#> $fit
#> $beta
#>               5.7260    5.3401     4.9802     4.6446    4.3315   4.0396
#> (Intercept) 19.66667 21.827014 23.8417604 25.7207200 27.473044 29.10727
#> am           0.00000  0.000000  0.0000000  0.0000000  0.000000  0.00000
#> carb         0.00000  0.000000  0.0000000  0.0000000  0.000000  0.00000
#> cyl          0.00000 -0.333583 -0.6446836 -0.9348171 -1.205396 -1.45774
#> disp         0.00000  0.000000  0.0000000  0.0000000  0.000000  0.00000
#> drat         0.00000  0.000000  0.0000000  0.0000000  0.000000  0.00000
#> gear         0.00000  0.000000  0.0000000  0.0000000  0.000000  0.00000
#> hp           0.00000  0.000000  0.0000000  0.0000000  0.000000  0.00000
#> qsec         0.00000  0.000000  0.0000000  0.0000000  0.000000  0.00000
#> vs           0.00000  0.000000  0.0000000  0.0000000  0.000000  0.00000
#> wt           0.00000  0.000000  0.0000000  0.0000000  0.000000  0.00000
#>                3.7673    3.5134    3.2766    3.0558    2.8499    2.6578
#> (Intercept) 30.631349 32.052713 33.378282 34.614512 35.767424 36.842634
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.693076 -1.912551 -2.117235 -2.308123 -2.486146 -2.652171
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                2.4787    2.3116    2.1558    2.0105    1.8750    1.7486
#> (Intercept) 37.845378 38.780541 39.652677 40.466033 41.036145 41.036145
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.807007 -2.951407 -3.086075 -3.211667 -3.299699 -3.299699
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.6308    1.5209    1.4184    1.3228    1.2336    1.1505
#> (Intercept) 41.036145 41.036145 41.036145 41.036145 41.036145 41.036145
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -3.299699 -3.299699 -3.299699 -3.299699 -3.299699 -3.299699
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                 1.0730    1.0006    0.9332    0.8703    0.8117      0.7569
#> (Intercept) 41.1771215 41.480877 41.480517 41.480284 41.480132 41.44995976
#> am           0.0000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> carb         0.0000000  0.000000  0.000000  0.000000  0.000000 -0.06322079
#> cyl         -2.8482590 -1.858310 -1.858136 -1.858024 -1.857951 -1.82292190
#> disp         0.0000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> drat         0.0000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> gear         0.0000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> hp           0.0000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> qsec         0.0000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> vs           0.0000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> wt          -0.9078721 -2.897122 -2.897348 -2.897495 -2.897590 -2.89966944
#>                 0.7059     0.6584     0.6140     0.5726     0.5340     0.4980
#> (Intercept) 41.4147963 41.3819142 41.3512442 41.3226410 41.2959655 41.2710879
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb        -0.1370465 -0.2059567 -0.2702244 -0.3301607 -0.3860576 -0.4381872
#> cyl         -1.7820512 -1.7438729 -1.7082651 -1.6750570 -1.6440870 -1.6152043
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.9020530 -2.9043156 -2.9064277 -2.9083977 -2.9102348 -2.9119481
#>                 0.4645     0.4332     0.4040    0.3767     0.3513     0.3277
#> (Intercept) 41.2478869 41.2264250 41.2062447 26.535123 26.5330136 26.5314784
#> am           0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> carb        -0.4868034 -0.5319719 -0.5742622 -0.925267 -0.9253885 -0.9254775
#> cyl         -1.5882682 -1.5632875 -1.5398571 -0.942479 -0.9422056 -0.9420053
#> disp         0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  3.051159  3.0515219  3.0517852
#> gear         0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> vs           0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> wt          -2.9135460 -2.9149717 -2.9163605 -2.531717 -2.5318806 -2.5320010
#>                 0.3056    0.2850    0.2658      0.2479      0.2312       0.2156
#> (Intercept) 26.5303611  3.769042 -7.185151 -25.2328435 -25.2386124 -60.93188452
#> am           0.0000000  0.000000  0.000000   0.0000000   0.0000000   0.00000000
#> carb        -0.9255426 -2.140370 -3.096834  -3.0613752  -3.0614460  -3.03381869
#> cyl         -0.9418586  0.000000  0.000000   0.0000000   0.0000000   0.00000000
#> disp         0.0000000  0.000000  0.000000   0.0000000   0.0000000   0.00000000
#> drat         3.0519764  3.687403  4.781981   5.0311086   5.0310417   4.87081613
#> gear         0.0000000  3.676845  5.725425   6.5459488   6.5464533   6.42224641
#> hp           0.0000000  0.000000  0.000000   0.0000000   0.0000000   0.04793971
#> qsec         0.0000000  0.000000  0.000000   0.8582219   0.8584945   3.14892219
#> vs           0.0000000  0.000000 -2.950288  -5.5908282  -5.5919201  -9.56583005
#> wt          -2.5320896 -1.109844  0.000000   0.0000000   0.0000000  -2.89606768
#>                   0.2011       0.1875       0.1749       0.1631       0.1521
#> (Intercept) -76.70151840 -76.71004952 -76.71825595 -76.72615003 -76.73374364
#> am           -2.73851399  -2.73912529  -2.73971342  -2.74027924  -2.74082360
#> carb         -3.47379938  -3.47408959  -3.47436881  -3.47463745  -3.47489592
#> cyl           0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> disp          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> drat          5.91581397   5.91616097   5.91649480   5.91681595   5.91712491
#> gear          8.28733466   8.28819613   8.28902496   8.28982238   8.29058957
#> hp            0.05806967   0.05807592   0.05808193   0.05808771   0.05809328
#> qsec          3.63909775   3.63938004   3.63965156   3.63991271   3.64016388
#> vs          -11.87573163 -11.87690560 -11.87803491 -11.87912127 -11.88016630
#> wt           -3.36540059  -3.36545830  -3.36551373  -3.36556697  -3.36561811
#>                   0.1418       0.1323       0.1234       0.1150       0.1073
#> (Intercept) -76.74104820 -76.74807470 -76.75483372 -76.76133543 -76.76758961
#> am           -2.74134730  -2.74185112  -2.74233583  -2.74280213  -2.74325073
#> carb         -3.47514460  -3.47538384  -3.47561402  -3.47583547  -3.47604851
#> cyl           0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> disp          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> drat          5.91742214   5.91770807   5.91798314   5.91824776   5.91850233
#> gear          8.29132768   8.29203779   8.29272097   8.29337822   8.29401054
#> hp            0.05809863   0.05810378   0.05810873   0.05811349   0.05811807
#> qsec          3.64040547   3.64063783   3.64086132   3.64107628   3.64128304
#> vs          -11.88117157 -11.88213859 -11.88306882 -11.88396365 -11.88482443
#> wt           -3.36566724  -3.36571444  -3.36575979  -3.36580335  -3.36584522
#>                   0.1001       0.0933       0.0870       0.0812       0.0757
#> (Intercept) -76.77360569 -76.77939271 -76.78495940 -76.79031413 -76.79546497
#> am           -2.74368229  -2.74409745  -2.74449685  -2.74488108  -2.74525070
#> carb         -3.47625348  -3.47645067  -3.47664037  -3.47682287  -3.47699845
#> cyl           0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> disp          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> drat          5.91874722   5.91898281   5.91920944   5.91942745   5.91963718
#> gear          8.29461886   8.29520409   8.29576710   8.29630872   8.29682978
#> hp            0.05812248   0.05812672   0.05813079   0.05813471   0.05813849
#> qsec          3.64148190   3.64167318   3.64185716   3.64203411   3.64220432
#> vs          -11.88565245 -11.88644896 -11.88721515 -11.88795218 -11.88866115
#> wt           -3.36588544  -3.36592410  -3.36596124  -3.36599694  -3.36603125
#>                   0.0706      0.0658       0.0614       0.0573      0.0534
#> (Intercept) -76.80041968 -76.8051857 -76.80977027 -76.81418024 -76.8184223
#> am           -2.74560628  -2.7459483  -2.74627741  -2.74659396  -2.7468985
#> carb         -3.47716735  -3.4773298  -3.47748617  -3.47763655  -3.4777812
#> cyl           0.00000000   0.0000000   0.00000000   0.00000000   0.0000000
#> disp          0.00000000   0.0000000   0.00000000   0.00000000   0.0000000
#> drat          5.91983893   5.9200330   5.92021970   5.92039930   5.9205721
#> gear          8.29733105   8.2978133   8.29827717   8.29872345   8.2991528
#> hp            0.05814211   0.0581456   0.05814896   0.05815219   0.0581553
#> qsec          3.64236803   3.6425255   3.64267696   3.64282265   3.6429628
#> vs          -11.88934314 -11.8899992 -11.89063022 -11.89123724 -11.8918212
#> wt           -3.36606422  -3.3660959  -3.36612638  -3.36615566  -3.3661838
#>                   0.0498       0.0464       0.0433       0.0404       0.0377
#> (Intercept) -76.82250277 -76.82642786 -76.83020347 -76.83383529 -76.83732879
#> am           -2.74719142  -2.74747322  -2.74774430  -2.74800507  -2.74825593
#> carb         -3.47792039  -3.47805427  -3.47818306  -3.47830696  -3.47842615
#> cyl           0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> disp          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> drat          5.92073826   5.92089813   5.92105192   5.92119986   5.92134217
#> gear          8.29956575   8.29996305   8.30034524   8.30071291   8.30106659
#> hp            0.05815828   0.05816116   0.05816392   0.05816658   0.05816914
#> qsec          3.64309756   3.64322720   3.64335190   3.64347184   3.64358721
#> vs          -11.89238285 -11.89292315 -11.89344288 -11.89394282 -11.89442372
#> wt           -3.36621086  -3.36623687  -3.36626188  -3.36628591  -3.36630902
#>                  0.0351       0.0328       0.0306       0.0285       0.0266
#> (Intercept) -76.8406892 -76.84392170 -76.84703104 -76.85002196 -76.85289897
#> am           -2.7484972  -2.74872938  -2.74895268  -2.74916748  -2.74937412
#> carb         -3.4785408  -3.47865110  -3.47875720  -3.47885927  -3.47895745
#> cyl           0.0000000   0.00000000   0.00000000   0.00000000   0.00000000
#> disp          0.0000000   0.00000000   0.00000000   0.00000000   0.00000000
#> drat          5.9214791   5.92161075   5.92173742   5.92185927   5.92197649
#> gear          8.3014068   8.30173411   8.30204895   8.30235182   8.30264317
#> hp            0.0581716   0.05817396   0.05817624   0.05817843   0.05818053
#> qsec          3.6436982   3.64380492   3.64390759   3.64400635   3.64410134
#> vs          -11.8948863 -11.89533129 -11.89575932 -11.89617106 -11.89656711
#> wt           -3.3663312  -3.36635260  -3.36637313  -3.36639288  -3.36641186
#>                   0.0248       0.0231       0.0216       0.0201       0.0188
#> (Intercept) -76.85566639 -76.85832840 -76.86088902 -76.86335211 -76.86572137
#> am           -2.74957289  -2.74976409  -2.74994802  -2.75012495  -2.75029515
#> carb         -3.47905190  -3.47914276  -3.47923016  -3.47931424  -3.47939511
#> cyl           0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> disp          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> drat          5.92208924   5.92219770   5.92230204   5.92240240   5.92249894
#> gear          8.30292343   8.30319303   8.30345238   8.30370185   8.30394184
#> hp            0.05818256   0.05818451   0.05818638   0.05818819   0.05818992
#> qsec          3.64419270   3.64428059   3.64436512   3.64444644   3.64452465
#> vs          -11.89694808 -11.89731454 -11.89766705 -11.89800613 -11.89833229
#> wt           -3.36643012  -3.36644767  -3.36646454  -3.36648077  -3.36649638
#>                   0.0175       0.0163       0.0152       0.0142       0.0132
#> (Intercept) -76.86800039 -76.87019260 -76.87230131 -76.87432969 -76.87628080
#> am           -2.75045887  -2.75061635  -2.75076785  -2.75091357  -2.75105375
#> carb         -3.47947291  -3.47954775  -3.47961974  -3.47968899  -3.47975561
#> cyl           0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> disp          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> drat          5.92259180   5.92268113   5.92276706   5.92284972   5.92292923
#> gear          8.30417268   8.30439475   8.30460836   8.30481384   8.30501151
#> hp            0.05819159   0.05819319   0.05819474   0.05819622   0.05819765
#> qsec          3.64459988   3.64467225   3.64474185   3.64480880   3.64487320
#> vs          -11.89864604 -11.89894783 -11.89923813 -11.89951738 -11.89978598
#> wt           -3.36651138  -3.36652581  -3.36653968  -3.36655303  -3.36656586
#>                   0.0123       0.0115       0.0107       0.0100       0.0093
#> (Intercept) -76.87815760 -76.87996290 -76.88169943 -76.88336981 -76.88497656
#> am           -2.75118859  -2.75131830  -2.75144307  -2.75156309  -2.75167854
#> carb         -3.47981969  -3.47988133  -3.47994063  -3.47999766  -3.48005253
#> cyl           0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> disp          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> drat          5.92300571   5.92307928   5.92315005   5.92321813   5.92328361
#> gear          8.30520164   8.30538454   8.30556048   8.30572972   8.30589252
#> hp            0.05819902   0.05820034   0.05820161   0.05820284   0.05820401
#> qsec          3.64493515   3.64499473   3.64505205   3.64510718   3.64516021
#> vs          -11.90004436 -11.90029290 -11.90053197 -11.90076193 -11.90098313
#> wt           -3.36657820  -3.36659006  -3.36660147  -3.36661245  -3.36662300
#>                   0.0087       0.0081       0.0076       0.0071       0.0066
#> (Intercept) -76.88652211 -76.88800878 -76.88943881 -76.89081437 -76.89213753
#> am           -2.75178959  -2.75189641  -2.75199917  -2.75209802  -2.75219310
#> carb         -3.48010531  -3.48015608  -3.48020491  -3.48025189  -3.48029708
#> cyl           0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> disp          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> drat          5.92334660   5.92340719   5.92346547   5.92352153   5.92357546
#> gear          8.30604912   8.30619976   8.30634466   8.30648404   8.30661812
#> hp            0.05820514   0.05820623   0.05820728   0.05820828   0.05820925
#> qsec          3.64521121   3.64526028   3.64530747   3.64535287   3.64539653
#> vs          -11.90119591 -11.90140058 -11.90159746 -11.90178683 -11.90196900
#> wt           -3.36663315  -3.36664291  -3.36665230  -3.36666133  -3.36667001
#>                   0.0061       0.0057
#> (Intercept) -76.89341028 -76.89463454
#> am           -2.75228456  -2.75237253
#> carb         -3.48034055  -3.48038236
#> cyl           0.00000000   0.00000000
#> disp          0.00000000   0.00000000
#> drat          5.92362733   5.92367723
#> gear          8.30674709   8.30687115
#> hp            0.05821018   0.05821108
#> qsec          3.64543853   3.64547893
#> vs          -11.90214422 -11.90231277
#> wt           -3.36667836  -3.36668639
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1   1 138  37   1   1   1   8   7   7   7   7   7   7
#>  [37]   7   6   6  44   1   1   1  57  60  48   1 108 125   1   1   1   1   1
#>  [55]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.726035563 5.340119930 4.980213720 4.644563983 4.331535916 4.039604892
#>   [7] 3.767349042 3.513442326 3.276648072 3.055812959 2.849861393 2.657790273
#>  [13] 2.478664104 2.311610439 2.155815632 2.010520874 1.875018496 1.748648525
#>  [19] 1.630795467 1.520885310 1.418382731 1.322788482 1.233636965 1.150493963
#>  [25] 1.072954520 1.000640977 0.933201124 0.870306491 0.811650747 0.756948203
#>  [31] 0.705932427 0.658354944 0.613984025 0.572603556 0.534011993 0.498021372
#>  [37] 0.464456398 0.433153592 0.403960489 0.376734904 0.351344233 0.327664807
#>  [43] 0.305581296 0.284986139 0.265779027 0.247866410 0.231161044 0.215581563
#>  [49] 0.201052087 0.187501850 0.174864852 0.163079547 0.152088531 0.141838273
#>  [55] 0.132278848 0.123363697 0.115049396 0.107295452 0.100064098 0.093320112
#>  [61] 0.087030649 0.081165075 0.075694820 0.070593243 0.065835494 0.061398402
#>  [67] 0.057260356 0.053401199 0.049802137 0.046445640 0.043315359 0.040396049
#>  [73] 0.037673490 0.035134423 0.032766481 0.030558130 0.028498614 0.026577903
#>  [79] 0.024786641 0.023116104 0.021558156 0.020105209 0.018750185 0.017486485
#>  [85] 0.016307955 0.015208853 0.014183827 0.013227885 0.012336370 0.011504940
#>  [91] 0.010729545 0.010006410 0.009332011 0.008703065 0.008116507 0.007569482
#>  [97] 0.007059324 0.006583549 0.006139840 0.005726036
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 24
#> 
#> $loss
#>   [1] 900.44667 768.26835 657.68154 565.57904 489.27826 426.46455 375.14202
#>   [8] 333.59088 300.33037 274.08662 253.76476 238.42473 227.26027 219.58075
#>  [15] 214.79539 212.39959 211.90952 211.90952 211.90952 211.90952 211.90952
#>  [22] 211.90952 211.90952 211.90952 174.31220 140.76230 140.76229 140.76229
#>  [29] 140.76229 137.18344 133.48698 130.50592 128.13405 126.27721 124.85442
#>  [36] 123.79621 123.04300 122.54531 122.25580 110.60601 110.60600 110.60600
#>  [43] 110.60599  90.46716  84.48461  77.20869  77.20867  47.26773  41.47970
#>  [50]  41.47967  41.47964  41.47962  41.47959  41.47957  41.47955  41.47953
#>  [57]  41.47951  41.47949  41.47948  41.47946  41.47945  41.47944  41.47943
#>  [64]  41.47942  41.47941  41.47940  41.47939  41.47938  41.47938  41.47937
#>  [71]  41.47936  41.47936  41.47935  41.47935  41.47934  41.47934  41.47933
#>  [78]  41.47933  41.47933  41.47932  41.47932  41.47932  41.47932  41.47931
#>  [85]  41.47931  41.47931  41.47931  41.47931  41.47931  41.47930  41.47930
#>  [92]  41.47930  41.47930  41.47930  41.47930  41.47930  41.47930  41.47930
#>  [99]  41.47930  41.47930
#> 
#> $linear.predictors
#>      5.7260   5.3401   4.9802   4.6446   4.3315   4.0396   3.7673   3.5134
#> 1  19.66667 19.82552 19.97366 20.11182 20.24066 20.36083 20.47289 20.57741
#> 2  19.66667 20.49268 21.26303 21.98145 22.65146 23.27631 23.85905 24.40251
#> 3  19.66667 19.15835 18.68429 18.24218 17.82987 17.44535 17.08674 16.75230
#> 4  19.66667 19.82552 19.97366 20.11182 20.24066 20.36083 20.47289 20.57741
#> 5  19.66667 19.15835 18.68429 18.24218 17.82987 17.44535 17.08674 16.75230
#> 6  19.66667 20.49268 21.26303 21.98145 22.65146 23.27631 23.85905 24.40251
#> 7  19.66667 19.82552 19.97366 20.11182 20.24066 20.36083 20.47289 20.57741
#> 8  19.66667 19.15835 18.68429 18.24218 17.82987 17.44535 17.08674 16.75230
#> 9  19.66667 19.15835 18.68429 18.24218 17.82987 17.44535 17.08674 16.75230
#> 10 19.66667 19.15835 18.68429 18.24218 17.82987 17.44535 17.08674 16.75230
#> 11 19.66667 19.15835 18.68429 18.24218 17.82987 17.44535 17.08674 16.75230
#> 12 19.66667 19.15835 18.68429 18.24218 17.82987 17.44535 17.08674 16.75230
#> 13 19.66667 20.49268 21.26303 21.98145 22.65146 23.27631 23.85905 24.40251
#> 14 19.66667 20.49268 21.26303 21.98145 22.65146 23.27631 23.85905 24.40251
#> 15 19.66667 19.15835 18.68429 18.24218 17.82987 17.44535 17.08674 16.75230
#> 16 19.66667 19.15835 18.68429 18.24218 17.82987 17.44535 17.08674 16.75230
#> 17 19.66667 19.15835 18.68429 18.24218 17.82987 17.44535 17.08674 16.75230
#> 18 19.66667 20.49268 21.26303 21.98145 22.65146 23.27631 23.85905 24.40251
#> 19 19.66667 20.49268 21.26303 21.98145 22.65146 23.27631 23.85905 24.40251
#> 20 19.66667 19.82552 19.97366 20.11182 20.24066 20.36083 20.47289 20.57741
#> 21 19.66667 19.15835 18.68429 18.24218 17.82987 17.44535 17.08674 16.75230
#>      3.2766   3.0558   2.8499   2.6578   2.4787   2.3116   2.1558   2.0105
#> 1  20.67487 20.76577 20.85055 20.92961 21.00334 21.07210 21.13623 21.19603
#> 2  24.90934 25.38202 25.82284 26.23395 26.61735 26.97491 27.30838 27.61937
#> 3  16.44040 16.14953 15.87825 15.62526 15.38932 15.16928 14.96408 14.77270
#> 4  20.67487 20.76577 20.85055 20.92961 21.00334 21.07210 21.13623 21.19603
#> 5  16.44040 16.14953 15.87825 15.62526 15.38932 15.16928 14.96408 14.77270
#> 6  24.90934 25.38202 25.82284 26.23395 26.61735 26.97491 27.30838 27.61937
#> 7  20.67487 20.76577 20.85055 20.92961 21.00334 21.07210 21.13623 21.19603
#> 8  16.44040 16.14953 15.87825 15.62526 15.38932 15.16928 14.96408 14.77270
#> 9  16.44040 16.14953 15.87825 15.62526 15.38932 15.16928 14.96408 14.77270
#> 10 16.44040 16.14953 15.87825 15.62526 15.38932 15.16928 14.96408 14.77270
#> 11 16.44040 16.14953 15.87825 15.62526 15.38932 15.16928 14.96408 14.77270
#> 12 16.44040 16.14953 15.87825 15.62526 15.38932 15.16928 14.96408 14.77270
#> 13 24.90934 25.38202 25.82284 26.23395 26.61735 26.97491 27.30838 27.61937
#> 14 24.90934 25.38202 25.82284 26.23395 26.61735 26.97491 27.30838 27.61937
#> 15 16.44040 16.14953 15.87825 15.62526 15.38932 15.16928 14.96408 14.77270
#> 16 16.44040 16.14953 15.87825 15.62526 15.38932 15.16928 14.96408 14.77270
#> 17 16.44040 16.14953 15.87825 15.62526 15.38932 15.16928 14.96408 14.77270
#> 18 24.90934 25.38202 25.82284 26.23395 26.61735 26.97491 27.30838 27.61937
#> 19 24.90934 25.38202 25.82284 26.23395 26.61735 26.97491 27.30838 27.61937
#> 20 20.67487 20.76577 20.85055 20.92961 21.00334 21.07210 21.13623 21.19603
#> 21 16.44040 16.14953 15.87825 15.62526 15.38932 15.16928 14.96408 14.77270
#>      1.8750   1.7486   1.6308   1.5209   1.4184   1.3228   1.2336   1.1505
#> 1  21.23795 21.23795 21.23795 21.23795 21.23795 21.23795 21.23795 21.23795
#> 2  27.83735 27.83735 27.83735 27.83735 27.83735 27.83735 27.83735 27.83735
#> 3  14.63855 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855
#> 4  21.23795 21.23795 21.23795 21.23795 21.23795 21.23795 21.23795 21.23795
#> 5  14.63855 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855
#> 6  27.83735 27.83735 27.83735 27.83735 27.83735 27.83735 27.83735 27.83735
#> 7  21.23795 21.23795 21.23795 21.23795 21.23795 21.23795 21.23795 21.23795
#> 8  14.63855 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855
#> 9  14.63855 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855
#> 10 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855
#> 11 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855
#> 12 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855
#> 13 27.83735 27.83735 27.83735 27.83735 27.83735 27.83735 27.83735 27.83735
#> 14 27.83735 27.83735 27.83735 27.83735 27.83735 27.83735 27.83735 27.83735
#> 15 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855
#> 16 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855
#> 17 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855
#> 18 27.83735 27.83735 27.83735 27.83735 27.83735 27.83735 27.83735 27.83735
#> 19 27.83735 27.83735 27.83735 27.83735 27.83735 27.83735 27.83735 27.83735
#> 20 21.23795 21.23795 21.23795 21.23795 21.23795 21.23795 21.23795 21.23795
#> 21 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855 14.63855
#>      1.0730   1.0006   0.9332   0.8703   0.8117   0.7569   0.7059   0.6584
#> 1  21.70894 22.74056 22.74065 22.74070 22.74074 22.66241 22.57092 22.48554
#> 2  27.67782 27.32631 27.32612 27.32600 27.32592 27.36782 27.41678 27.46245
#> 3  15.26797 16.64830 16.64855 16.64871 16.64882 16.76528 16.90123 17.02817
#> 4  20.94633 20.30698 20.30687 20.30681 20.30677 20.41635 20.54434 20.66379
#> 5  15.14995 16.27167 16.27189 16.27204 16.27213 16.26188 16.24987 16.23870
#> 6  26.88797 24.80582 24.80543 24.80518 24.80502 24.78189 24.75495 24.72974
#> 7  20.96449 20.36492 20.36482 20.36476 20.36472 20.28468 20.19124 20.10400
#> 8  14.69601 14.82311 14.82322 14.82329 14.82333 14.87527 14.93589 14.99250
#> 9  15.00469 15.80813 15.80832 15.80844 15.80852 15.86116 15.92259 15.97996
#> 10 13.62472 11.40451 11.40435 11.40424 11.40418 11.39044 11.37442 11.35945
#> 11 13.46675 10.90041 10.90021 10.90008 10.90000 10.88589 10.86947 10.85410
#> 12 13.53847 11.12928 11.12910 11.12898 11.12891 11.11497 11.09873 11.08354
#> 13 27.78677 27.67397 27.67381 27.67370 27.67363 27.71578 27.76503 27.81097
#> 14 28.11814 28.73142 28.73134 28.73129 28.73125 28.77416 28.82428 28.87105
#> 15 15.19534 16.41653 16.41676 16.41691 16.41701 16.53331 16.66907 16.79583
#> 16 14.90482 15.48945 15.48961 15.48971 15.48978 15.47897 15.46632 15.45453
#> 17 14.90028 15.47496 15.47512 15.47523 15.47529 15.59091 15.72590 15.85192
#> 18 28.02735 28.44171 28.44160 28.44154 28.44149 28.48419 28.53407 28.58062
#> 19 28.41048 29.66429 29.66428 29.66428 29.66428 29.64463 29.62169 29.60028
#> 20 21.57276 22.30599 22.30604 22.30608 22.30610 22.10102 21.86152 21.63798
#> 21 15.14995 16.27167 16.27189 16.27204 16.27213 16.00900 15.70169 15.41487
#>      0.6140   0.5726   0.5340   0.4980   0.4645   0.4332   0.4040   0.3767
#> 1  22.40592 22.33165 22.26240 22.19781 22.13757 22.08159 22.02919 22.44560
#> 2  27.50505 27.54477 27.58182 27.61636 27.64858 27.67857 27.70660 27.71332
#> 3  17.14656 17.25698 17.35995 17.45598 17.54554 17.62868 17.70658 18.04680
#> 4  20.77519 20.87908 20.97597 21.06633 21.15061 21.22893 21.30223 19.61644
#> 5  16.22828 16.21856 16.20950 16.20105 16.19317 16.18579 16.17893 16.05021
#> 6  24.70623 24.68430 24.66385 24.64478 24.62700 24.61057 24.59510 24.09727
#> 7  20.02264 19.94677 19.87601 19.81001 19.74847 19.69131 19.63777 20.43062
#> 8  15.04529 15.09452 15.14044 15.18326 15.22320 15.26027 15.29501 15.28246
#> 9  16.03348 16.08338 16.12992 16.17333 16.21380 16.25137 16.28658 16.14324
#> 10 11.34548 11.33245 11.32031 11.30898 11.29841 11.28864 11.27945 10.94260
#> 11 10.83976 10.82639 10.81393 10.80230 10.79145 10.78143 10.77200 10.71567
#> 12 11.06937 11.05616 11.04383 11.03234 11.02162 11.01171 11.00239 11.61744
#> 13 27.85382 27.89378 27.93104 27.96580 27.99821 28.02837 28.05656 28.71889
#> 14 28.91466 28.95534 28.99328 29.02866 29.06165 29.09233 29.12103 30.07013
#> 15 16.91405 17.02430 17.12713 17.22302 17.31245 17.39548 17.47327 16.65431
#> 16 15.44354 15.43330 15.42374 15.41482 15.40651 15.39875 15.39151 16.95325
#> 17 15.96946 16.07907 16.18130 16.27664 16.36555 16.44812 16.52546 16.80787
#> 18 28.62402 28.66450 28.70226 28.73746 28.77030 28.80083 28.82940 29.38980
#> 19 29.58031 29.56169 29.54432 29.52812 29.51301 29.49898 29.48584 28.58706
#> 20 21.42950 21.23507 21.05375 20.88464 20.72693 20.58040 20.44321 19.36099
#> 21 15.14738 14.89792 14.66527 14.44830 14.24595 14.05790 13.88188 13.35603
#>      0.3513   0.3277   0.3056   0.2850   0.2658   0.2479   0.2312   0.2156
#> 1  22.44563 22.44566 22.44567 21.38802 21.97894 22.45311 22.45330 21.13494
#> 2  27.71320 27.71311 27.71305 27.95772 28.08005 27.64003 27.63993 27.25105
#> 3  18.04722 18.04752 18.04774 18.31629 18.86070 18.73718 18.73721 19.63192
#> 4  19.61628 19.61617 19.61609 18.99638 17.14227 16.99191 16.99182 17.86314
#> 5  16.05039 16.05051 16.05061 14.11252 12.95395 11.90359 11.90316 13.12009
#> 6  24.09683 24.09651 24.09627 24.26180 24.21810 24.96660 24.96682 23.80920
#> 7  20.43052 20.43045 20.43040 20.55170 19.12429 19.05696 19.05673 17.59834
#> 8  15.28262 15.28274 15.28283 15.18173 15.38130 15.59944 15.59951 15.82020
#> 9  16.14346 16.14362 16.14374 15.55908 15.38130 15.77109 15.77121 17.43465
#> 10 10.94240 10.94225 10.94214 11.21551 11.61499 12.33148 12.33165 11.71198
#> 11 10.71546 10.71531 10.71520 11.28051 11.94973 12.54634 12.54646 11.52459
#> 12 11.61733 11.61725 11.61719 12.21629 13.04959 13.36021 13.36020 12.33319
#> 13 28.71887 28.71886 28.71885 28.93900 29.17991 29.53525 29.53537 30.13257
#> 14 30.07022 30.07029 30.07034 29.86033 29.84939 30.60864 30.60887 33.17765
#> 15 16.65457 16.65476 16.65491 16.78942 16.99572 16.64631 16.64633 15.82979
#> 16 16.95357 16.95380 16.95397 15.73031 15.44058 14.15074 14.15015 13.51694
#> 17 16.80820 16.80844 16.80861 17.60869 18.52596 18.41075 18.41080 18.21252
#> 18 29.38982 29.38984 29.38985 29.23311 29.17991 29.04606 29.04603 29.10514
#> 19 28.58692 28.58681 28.58674 30.09484 30.32609 29.25455 29.25443 28.16108
#> 20 19.36065 19.36040 19.36022 19.58518 20.17174 20.64370 20.64401 19.78443
#> 21 13.35583 13.35569 13.35559 14.12157 13.59551 13.34606 13.34599 15.84659
#>      0.2011   0.1875   0.1749   0.1631   0.1521   0.1418   0.1323   0.1234
#> 1  20.35565 20.35532 20.35502 20.35472 20.35443 20.35416 20.35389 20.35364
#> 2  26.45202 26.45189 26.45177 26.45166 26.45155 26.45144 26.45134 26.45124
#> 3  20.37036 20.37062 20.37088 20.37113 20.37136 20.37159 20.37181 20.37202
#> 4  17.17419 17.17390 17.17362 17.17335 17.17310 17.17285 17.17261 17.17238
#> 5  13.11095 13.11075 13.11056 13.11038 13.11020 13.11004 13.10987 13.10972
#> 6  25.10049 25.10078 25.10106 25.10133 25.10159 25.10184 25.10207 25.10230
#> 7  18.21142 18.21127 18.21112 18.21098 18.21084 18.21071 18.21058 18.21046
#> 8  15.97630 15.97635 15.97640 15.97644 15.97649 15.97653 15.97657 15.97661
#> 9  17.84835 17.84848 17.84860 17.84872 17.84883 17.84894 17.84904 17.84914
#> 10 11.26553 11.26549 11.26546 11.26542 11.26539 11.26536 11.26533 11.26530
#> 11 11.09250 11.09249 11.09249 11.09248 11.09248 11.09247 11.09247 11.09247
#> 12 12.13441 12.13447 12.13453 12.13458 12.13463 12.13469 12.13474 12.13478
#> 13 29.77825 29.77828 29.77832 29.77835 29.77839 29.77842 29.77845 29.77848
#> 14 33.34157 33.34180 33.34201 33.34221 33.34241 33.34260 33.34278 33.34296
#> 15 15.79635 15.79628 15.79621 15.79614 15.79608 15.79602 15.79596 15.79590
#> 16 13.71370 13.71355 13.71340 13.71326 13.71312 13.71299 13.71287 13.71274
#> 17 18.70244 18.70266 18.70288 18.70309 18.70329 18.70349 18.70368 18.70386
#> 18 28.59579 28.59568 28.59558 28.59548 28.59538 28.59529 28.59520 28.59511
#> 19 28.44670 28.44681 28.44692 28.44702 28.44711 28.44721 28.44730 28.44738
#> 20 19.81514 19.81513 19.81512 19.81511 19.81510 19.81509 19.81508 19.81507
#> 21 15.71791 15.71799 15.71807 15.71815 15.71822 15.71829 15.71836 15.71842
#>      0.1150   0.1073   0.1001   0.0933   0.0870   0.0812   0.0757   0.0706
#> 1  20.35339 20.35316 20.35293 20.35271 20.35251 20.35230 20.35211 20.35192
#> 2  26.45115 26.45106 26.45097 26.45089 26.45081 26.45073 26.45066 26.45058
#> 3  20.37222 20.37242 20.37261 20.37279 20.37296 20.37313 20.37329 20.37344
#> 4  17.17216 17.17195 17.17175 17.17155 17.17136 17.17118 17.17101 17.17084
#> 5  13.10957 13.10942 13.10929 13.10915 13.10902 13.10890 13.10878 13.10867
#> 6  25.10253 25.10274 25.10294 25.10314 25.10333 25.10351 25.10369 25.10386
#> 7  18.21034 18.21023 18.21012 18.21001 18.20991 18.20982 18.20972 18.20964
#> 8  15.97665 15.97669 15.97672 15.97676 15.97679 15.97682 15.97685 15.97688
#> 9  17.84924 17.84933 17.84942 17.84950 17.84959 17.84967 17.84974 17.84982
#> 10 11.26527 11.26525 11.26522 11.26520 11.26517 11.26515 11.26513 11.26511
#> 11 11.09246 11.09246 11.09246 11.09245 11.09245 11.09245 11.09244 11.09244
#> 12 12.13483 12.13487 12.13492 12.13496 12.13500 12.13504 12.13507 12.13511
#> 13 29.77851 29.77853 29.77856 29.77858 29.77861 29.77863 29.77865 29.77867
#> 14 33.34313 33.34329 33.34345 33.34360 33.34374 33.34388 33.34401 33.34414
#> 15 15.79584 15.79579 15.79574 15.79569 15.79564 15.79560 15.79556 15.79551
#> 16 13.71263 13.71252 13.71241 13.71230 13.71220 13.71211 13.71202 13.71193
#> 17 18.70403 18.70420 18.70436 18.70451 18.70466 18.70480 18.70494 18.70507
#> 18 28.59503 28.59495 28.59487 28.59480 28.59473 28.59466 28.59459 28.59453
#> 19 28.44747 28.44755 28.44762 28.44770 28.44777 28.44784 28.44790 28.44797
#> 20 19.81507 19.81506 19.81505 19.81504 19.81504 19.81503 19.81502 19.81502
#> 21 15.71848 15.71854 15.71860 15.71865 15.71871 15.71876 15.71881 15.71885
#>      0.0658   0.0614   0.0573   0.0534   0.0498   0.0464   0.0433   0.0404
#> 1  20.35174 20.35157 20.35140 20.35124 20.35109 20.35094 20.35080 20.35066
#> 2  26.45052 26.45045 26.45039 26.45032 26.45026 26.45021 26.45015 26.45010
#> 3  20.37359 20.37374 20.37387 20.37401 20.37413 20.37426 20.37437 20.37449
#> 4  17.17068 17.17052 17.17037 17.17023 17.17009 17.16996 17.16983 17.16971
#> 5  13.10856 13.10845 13.10835 13.10825 13.10815 13.10806 13.10798 13.10789
#> 6  25.10402 25.10418 25.10433 25.10447 25.10461 25.10475 25.10488 25.10500
#> 7  18.20955 18.20947 18.20939 18.20931 18.20924 18.20917 18.20910 18.20903
#> 8  15.97691 15.97694 15.97696 15.97699 15.97701 15.97703 15.97706 15.97708
#> 9  17.84989 17.84995 17.85002 17.85008 17.85014 17.85020 17.85025 17.85031
#> 10 11.26509 11.26507 11.26505 11.26503 11.26501 11.26499 11.26498 11.26496
#> 11 11.09244 11.09244 11.09243 11.09243 11.09243 11.09243 11.09243 11.09242
#> 12 12.13514 12.13517 12.13521 12.13524 12.13527 12.13529 12.13532 12.13535
#> 13 29.77870 29.77871 29.77873 29.77875 29.77877 29.77879 29.77880 29.77882
#> 14 33.34427 33.34438 33.34450 33.34461 33.34471 33.34482 33.34491 33.34501
#> 15 15.79547 15.79543 15.79540 15.79536 15.79533 15.79529 15.79526 15.79523
#> 16 13.71184 13.71176 13.71168 13.71160 13.71153 13.71146 13.71139 13.71133
#> 17 18.70520 18.70532 18.70544 18.70555 18.70566 18.70576 18.70587 18.70596
#> 18 28.59447 28.59441 28.59436 28.59430 28.59425 28.59420 28.59415 28.59411
#> 19 28.44803 28.44809 28.44814 28.44820 28.44825 28.44830 28.44835 28.44840
#> 20 19.81501 19.81501 19.81500 19.81500 19.81499 19.81499 19.81498 19.81498
#> 21 15.71890 15.71894 15.71898 15.71902 15.71906 15.71910 15.71913 15.71917
#>      0.0377   0.0351   0.0328   0.0306   0.0285   0.0266   0.0248   0.0231
#> 1  20.35053 20.35041 20.35028 20.35017 20.35005 20.34995 20.34984 20.34974
#> 2  26.45005 26.45000 26.44995 26.44991 26.44987 26.44982 26.44978 26.44975
#> 3  20.37460 20.37470 20.37480 20.37490 20.37499 20.37508 20.37517 20.37525
#> 4  17.16959 17.16948 17.16937 17.16926 17.16916 17.16906 17.16897 17.16888
#> 5  13.10781 13.10773 13.10766 13.10759 13.10752 13.10745 13.10739 13.10733
#> 6  25.10512 25.10523 25.10534 25.10545 25.10555 25.10565 25.10575 25.10584
#> 7  18.20897 18.20891 18.20885 18.20879 18.20874 18.20869 18.20864 18.20859
#> 8  15.97710 15.97712 15.97714 15.97715 15.97717 15.97719 15.97721 15.97722
#> 9  17.85036 17.85041 17.85046 17.85050 17.85055 17.85059 17.85063 17.85067
#> 10 11.26495 11.26493 11.26492 11.26491 11.26489 11.26488 11.26487 11.26486
#> 11 11.09242 11.09242 11.09242 11.09242 11.09241 11.09241 11.09241 11.09241
#> 12 12.13537 12.13539 12.13542 12.13544 12.13546 12.13548 12.13550 12.13552
#> 13 29.77883 29.77885 29.77886 29.77888 29.77889 29.77890 29.77891 29.77892
#> 14 33.34510 33.34519 33.34527 33.34535 33.34543 33.34550 33.34557 33.34564
#> 15 15.79520 15.79517 15.79515 15.79512 15.79510 15.79507 15.79505 15.79502
#> 16 13.71126 13.71120 13.71115 13.71109 13.71104 13.71098 13.71093 13.71089
#> 17 18.70606 18.70615 18.70623 18.70631 18.70639 18.70647 18.70654 18.70662
#> 18 28.59406 28.59402 28.59398 28.59394 28.59390 28.59386 28.59383 28.59379
#> 19 28.44844 28.44849 28.44853 28.44857 28.44861 28.44864 28.44868 28.44871
#> 20 19.81497 19.81497 19.81497 19.81496 19.81496 19.81496 19.81495 19.81495
#> 21 15.71920 15.71923 15.71926 15.71929 15.71932 15.71935 15.71937 15.71940
#>      0.0216   0.0201   0.0188   0.0175   0.0163   0.0152   0.0142   0.0132
#> 1  20.34964 20.34955 20.34946 20.34938 20.34929 20.34921 20.34914 20.34906
#> 2  26.44971 26.44967 26.44964 26.44961 26.44957 26.44954 26.44951 26.44949
#> 3  20.37533 20.37541 20.37548 20.37555 20.37562 20.37569 20.37575 20.37581
#> 4  17.16879 17.16871 17.16863 17.16855 17.16847 17.16840 17.16833 17.16827
#> 5  13.10727 13.10721 13.10715 13.10710 13.10705 13.10700 13.10695 13.10691
#> 6  25.10592 25.10601 25.10609 25.10617 25.10624 25.10631 25.10638 25.10645
#> 7  18.20854 18.20850 18.20846 18.20842 18.20838 18.20834 18.20830 18.20827
#> 8  15.97724 15.97725 15.97726 15.97728 15.97729 15.97730 15.97732 15.97733
#> 9  17.85071 17.85074 17.85078 17.85081 17.85084 17.85088 17.85091 17.85093
#> 10 11.26485 11.26484 11.26483 11.26482 11.26481 11.26480 11.26479 11.26478
#> 11 11.09241 11.09241 11.09241 11.09240 11.09240 11.09240 11.09240 11.09240
#> 12 12.13554 12.13556 12.13557 12.13559 12.13560 12.13562 12.13563 12.13565
#> 13 29.77894 29.77895 29.77896 29.77897 29.77898 29.77899 29.77899 29.77900
#> 14 33.34571 33.34577 33.34583 33.34589 33.34595 33.34600 33.34606 33.34611
#> 15 15.79500 15.79498 15.79496 15.79494 15.79492 15.79491 15.79489 15.79487
#> 16 13.71084 13.71080 13.71075 13.71071 13.71067 13.71064 13.71060 13.71056
#> 17 18.70668 18.70675 18.70681 18.70687 18.70693 18.70699 18.70704 18.70710
#> 18 28.59376 28.59373 28.59370 28.59367 28.59364 28.59362 28.59359 28.59356
#> 19 28.44875 28.44878 28.44881 28.44884 28.44887 28.44889 28.44892 28.44895
#> 20 19.81495 19.81494 19.81494 19.81494 19.81494 19.81493 19.81493 19.81493
#> 21 15.71942 15.71945 15.71947 15.71949 15.71951 15.71953 15.71955 15.71957
#>      0.0123   0.0115   0.0107   0.0100   0.0093   0.0087   0.0081   0.0076
#> 1  20.34899 20.34893 20.34886 20.34880 20.34874 20.34868 20.34862 20.34857
#> 2  26.44946 26.44943 26.44941 26.44938 26.44936 26.44934 26.44932 26.44930
#> 3  20.37587 20.37593 20.37598 20.37603 20.37608 20.37613 20.37618 20.37622
#> 4  17.16820 17.16814 17.16808 17.16803 17.16797 17.16792 17.16787 17.16782
#> 5  13.10687 13.10682 13.10678 13.10674 13.10671 13.10667 13.10664 13.10660
#> 6  25.10651 25.10658 25.10664 25.10669 25.10675 25.10680 25.10685 25.10690
#> 7  18.20823 18.20820 18.20817 18.20814 18.20811 18.20808 18.20806 18.20803
#> 8  15.97734 15.97735 15.97736 15.97737 15.97738 15.97739 15.97740 15.97740
#> 9  17.85096 17.85099 17.85101 17.85104 17.85106 17.85108 17.85111 17.85113
#> 10 11.26477 11.26477 11.26476 11.26475 11.26474 11.26474 11.26473 11.26472
#> 11 11.09240 11.09240 11.09240 11.09240 11.09240 11.09240 11.09239 11.09239
#> 12 12.13566 12.13567 12.13569 12.13570 12.13571 12.13572 12.13573 12.13574
#> 13 29.77901 29.77902 29.77903 29.77903 29.77904 29.77905 29.77905 29.77906
#> 14 33.34616 33.34620 33.34625 33.34629 33.34633 33.34637 33.34641 33.34645
#> 15 15.79486 15.79484 15.79483 15.79481 15.79480 15.79479 15.79477 15.79476
#> 16 13.71053 13.71050 13.71047 13.71044 13.71041 13.71038 13.71035 13.71033
#> 17 18.70715 18.70719 18.70724 18.70728 18.70733 18.70737 18.70741 18.70745
#> 18 28.59354 28.59352 28.59350 28.59347 28.59345 28.59343 28.59341 28.59340
#> 19 28.44897 28.44899 28.44902 28.44904 28.44906 28.44908 28.44910 28.44912
#> 20 19.81493 19.81492 19.81492 19.81492 19.81492 19.81492 19.81491 19.81491
#> 21 15.71959 15.71960 15.71962 15.71964 15.71965 15.71966 15.71968 15.71969
#>      0.0071   0.0066   0.0061   0.0057
#> 1  20.34852 20.34847 20.34842 20.34837
#> 2  26.44928 26.44926 26.44924 26.44922
#> 3  20.37627 20.37631 20.37635 20.37639
#> 4  17.16778 17.16773 17.16769 17.16765
#> 5  13.10657 13.10654 13.10651 13.10648
#> 6  25.10695 25.10699 25.10704 25.10708
#> 7  18.20800 18.20798 18.20796 18.20794
#> 8  15.97741 15.97742 15.97743 15.97744
#> 9  17.85115 17.85117 17.85119 17.85120
#> 10 11.26472 11.26471 11.26471 11.26470
#> 11 11.09239 11.09239 11.09239 11.09239
#> 12 12.13575 12.13576 12.13577 12.13578
#> 13 29.77907 29.77907 29.77908 29.77908
#> 14 33.34648 33.34652 33.34655 33.34658
#> 15 15.79475 15.79474 15.79473 15.79472
#> 16 13.71030 13.71028 13.71026 13.71023
#> 17 18.70748 18.70752 18.70755 18.70759
#> 18 28.59338 28.59336 28.59335 28.59333
#> 19 28.44913 28.44915 28.44917 28.44918
#> 20 19.81491 19.81491 19.81491 19.81491
#> 21 15.71970 15.71972 15.71973 15.71974
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 22.8 18.7 18.1 14.3 24.4 17.8 16.4 17.3 10.4 10.4 14.7 32.4 33.9 15.5
#> [16] 13.3 19.2 27.3 30.4 19.7 15.0
#> 
#> $X
#>               am       carb        cyl       disp       drat       gear
#>  [1,]  1.2747549  0.5640761 -0.2744106 -0.6951287  0.9734722  0.5275044
#>  [2,]  1.2747549 -1.1281521 -1.4269354 -1.0886794  0.8628504  0.5275044
#>  [3,] -0.7844645 -0.5640761  0.8781141  0.8185278 -0.6858554 -0.8571946
#>  [4,] -0.7844645 -1.1281521 -0.2744106 -0.2031904 -1.5487058 -0.8571946
#>  [5,] -0.7844645  0.5640761  0.8781141  0.8185278 -0.5531092 -0.8571946
#>  [6,] -0.7844645 -0.5640761 -1.4269354 -0.7957869  0.5088605  0.5275044
#>  [7,] -0.7844645  0.5640761 -0.2744106 -0.6376098  1.0177209  0.5275044
#>  [8,] -0.7844645  0.0000000  0.8781141  0.1812784 -0.8628504 -0.8571946
#>  [9,] -0.7844645  0.0000000  0.8781141  0.1812784 -0.8628504 -0.8571946
#> [10,] -0.7844645  0.5640761  0.8781141  1.6661754 -1.1725915 -0.8571946
#> [11,] -0.7844645  0.5640761  0.8781141  1.5753560 -1.0177209 -0.8571946
#> [12,] -0.7844645  0.5640761  0.8781141  1.4239904 -0.5088605 -0.8571946
#> [13,]  1.2747549 -1.1281521 -1.4269354 -1.3104301  1.3717108  0.5275044
#> [14,]  1.2747549 -1.1281521 -1.4269354 -1.3679490  1.6814520  0.5275044
#> [15,] -0.7844645 -0.5640761  0.8781141  0.5006599 -1.5487058 -0.8571946
#> [16,] -0.7844645  0.5640761  0.8781141  0.7428449  0.5973579 -0.8571946
#> [17,] -0.7844645 -0.5640761  0.8781141  1.1212591 -0.8407260 -0.8571946
#> [18,]  1.2747549 -1.1281521 -1.4269354 -1.3081596  1.3717108  0.5275044
#> [19,]  1.2747549 -0.5640761 -1.4269354 -1.1863103  0.6858554  1.9122034
#> [20,]  1.2747549  1.6922282 -0.2744106 -0.8086530  0.3539899  1.9122034
#> [21,]  1.2747549  2.8203804  0.8781141  0.3719991  0.1769949  1.9122034
#>               hp         qsec         vs          wt
#>  [1,] -0.6710353 -0.744622539 -0.7844645 -0.70486573
#>  [2,] -0.9098783  0.644047748  1.2747549 -0.98473301
#>  [3,]  0.2421882 -0.382922371 -0.7844645  0.06010483
#>  [4,] -0.7412832  1.683935730  1.2747549  0.07876265
#>  [5,]  1.2256596 -1.145076296 -0.7844645  0.18138065
#>  [6,] -1.3454157  1.541839235  1.2747549 -0.17311790
#>  [7,] -0.4883906  0.831356763  1.2747549  0.06010483
#>  [8,]  0.3124362 -0.137482972 -0.7844645  0.64782611
#>  [9,]  0.3124362 -0.008304341 -0.7844645  0.33064320
#> [10,]  0.6636760  0.237135059 -0.7844645  1.74863740
#> [11,]  0.8041719  0.133792154 -0.7844645  1.91096042
#> [12,]  1.0149158 -0.124565109 -0.7844645  1.83726204
#> [13,] -1.2892173  1.199515862  1.2747549 -1.09667992
#> [14,] -1.3032669  1.477249920  1.2747549 -1.43718510
#> [15,] -0.1090516 -0.479806345 -0.7844645  0.13473610
#> [16,]  1.2256596 -1.422810354 -0.7844645  0.43326120
#> [17,]  0.2421882 -0.363545577 -0.7844645  0.43792565
#> [18,] -1.2892173  0.831356763  1.2747549 -1.34389601
#> [19,] -0.6288865 -0.460429550  1.2747549 -1.73757598
#> [20,]  0.2421882 -1.364679970 -0.7844645 -0.56493209
#> [21,]  2.4901229 -1.945983811 -0.7844645  0.18138065
#> attr(,"center")
#>  [1]   0.3809524   3.0000000   6.4761905 251.8476190   3.4600000   3.6190476
#>  [7] 157.7619048  17.6128571   0.3809524   3.3755714
#> attr(,"scale")
#>  [1]   0.4856209   1.7728105   1.7353207 132.1303744   0.4519903   0.7221786
#>  [7]  71.1764439   1.5482437   0.4856209   1.0719367
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 72
#> 
#> $lambda.min
#> [1] 0.04039605
#> 
#> $null.dev
#> [1] 42.87841
#> 
#> $Bias
#> [1] 3.08367
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 69.59849