Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 40.458322 37.851384 32.953054 27.573324 23.216618 19.727962 16.974034
#>   [8] 14.840242 13.228193 12.053496 11.243862 10.737449 10.481433 10.430770
#>  [15] 10.547118 10.797915 11.001404 11.234387 11.161655 10.801483 10.343715
#>  [22]  9.893550  9.202721  8.546394  6.909911  6.697260  6.524108  6.387281
#>  [29]  6.296626  6.229935  6.183226  6.152098  6.124863  6.046133  6.073082
#>  [36]  6.203857  6.390302  6.543570  6.905039  6.946128  6.981693  7.063640
#>  [43]  7.907139  7.697578  7.731619  7.889274 10.398558 10.146936 10.147195
#>  [50]  9.815832  9.136178  9.669933  9.557341  9.557774  9.080521  9.395323
#>  [57]  9.234018  9.466652  8.807899  8.303847  8.304292  8.304701  8.305078
#>  [64]  8.305425 10.826120 10.481754 10.914599 10.915385 10.916143 10.916876
#>  [71] 10.917584 10.918268 10.918930 10.832266 10.832967 10.917656 10.918286
#>  [78] 10.918951 10.885700 10.914614 10.915289 10.915948 10.916593 10.917223
#>  [85] 10.917840 10.918443 10.919032 10.919609 10.920173 10.955025 11.014731
#>  [92] 11.015402 11.016067 11.035936 11.040772 11.041569 11.042352 11.043120
#>  [99] 11.043875 11.044617
#> 
#> $cvse
#>   [1] 10.893839 10.791534  9.526779  8.084401  6.950957  6.055966  5.338852
#>   [8]  4.750663  4.254954  3.827497  3.455136  3.134232  2.868928  2.668909
#>  [15]  2.545905  2.508725  2.530768  2.608373  2.616303  2.564572  2.482940
#>  [22]  2.411588  2.320807  2.251590  1.997616  1.934679  1.878484  1.829069
#>  [29]  1.785385  1.749299  1.720749  1.699322  1.683792  1.671713  1.675845
#>  [36]  1.701878  1.748204  1.791977  1.833839  1.846641  1.857392  1.992718
#>  [43]  1.940143  1.972863  1.980982  2.020631  4.106903  4.092149  4.092362
#>  [50]  4.116374  4.064323  4.192801  4.204098  4.204277  3.600556  3.509748
#>  [57]  3.509633  3.462060  3.474555  3.479723  3.479836  3.479938  3.480028
#>  [64]  3.480110  5.369475  5.379758  5.378030  5.378549  5.379060  5.379564
#>  [71]  5.380062  5.380552  5.381036  5.361706  5.362201  5.358549  5.359028
#>  [78]  5.359500  5.367006  5.367648  5.368098  5.368540  5.368975  5.369403
#>  [85]  5.369824  5.370238  5.370646  5.371047  5.371442  5.380992  5.376343
#>  [92]  5.376722  5.377094  5.376454  5.376563  5.376910  5.377253  5.377590
#>  [99]  5.377922  5.378249
#> 
#> $fold
#>  [1]  9  6  6  7  2  4  5  8  3  1  4 10  1  9  5  3  8  7  1 10  2
#> 
#> $lambda
#>   [1] 5.535041084 5.161997840 4.814096462 4.489642473 4.187055597 3.904862063
#>   [7] 3.641687429 3.396249885 3.167354009 2.953884950 2.754802990 2.569138489
#>  [13] 2.395987153 2.234505638 2.083907437 1.943459050 1.812476415 1.690321571
#>  [19] 1.576399554 1.470155500 1.371071940 1.278666280 1.192488452 1.112118722
#>  [25] 1.037165642 0.967264149 0.902073783 0.841277029 0.784577773 0.731699857
#>  [31] 0.682385735 0.636395221 0.593504313 0.553504108 0.516199784 0.481409646
#>  [37] 0.448964247 0.418705560 0.390486206 0.364168743 0.339624988 0.316735401
#>  [43] 0.295388495 0.275480299 0.256913849 0.239598715 0.223450564 0.208390744
#>  [49] 0.194345905 0.181247641 0.169032157 0.157639955 0.147015550 0.137107194
#>  [55] 0.127866628 0.119248845 0.111211872 0.103716564 0.096726415 0.090207378
#>  [61] 0.084127703 0.078457777 0.073169986 0.068238574 0.063639522 0.059350431
#>  [67] 0.055350411 0.051619978 0.048140965 0.044896425 0.041870556 0.039048621
#>  [73] 0.036416874 0.033962499 0.031673540 0.029538849 0.027548030 0.025691385
#>  [79] 0.023959872 0.022345056 0.020839074 0.019434591 0.018124764 0.016903216
#>  [85] 0.015763996 0.014701555 0.013710719 0.012786663 0.011924885 0.011121187
#>  [91] 0.010371656 0.009672641 0.009020738 0.008412770 0.007845778 0.007316999
#>  [97] 0.006823857 0.006363952 0.005935043 0.005535041
#> 
#> $fit
#> $beta
#>               5.5350    5.1620    4.8141   4.4896    4.1871    3.9049    3.6417
#> (Intercept) 20.54286 22.475742 24.278357 25.95948 27.527305 28.989461 30.353073
#> am           0.00000  0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> carb         0.00000  0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> disp         0.00000  0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> drat         0.00000  0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> gear         0.00000  0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> hp           0.00000  0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> vs           0.00000  0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> wt           0.00000 -0.621535 -1.201181 -1.74176 -2.245906 -2.716075 -3.154555
#>                3.3962    3.1674    2.9539    2.7548    2.5691    2.3960
#> (Intercept) 31.624783 32.810783 33.916851 34.948373 35.910375 36.807541
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -3.563484 -3.944852 -4.300517 -4.632212 -4.941551 -5.230042
#>               2.2345    2.0839    1.9435   1.8125      1.6903     1.5764
#> (Intercept) 37.64424 38.424549 39.152268 39.66235 38.91461858 37.2024298
#> am           0.00000  0.000000  0.000000  0.00000  0.00000000  0.0000000
#> carb         0.00000  0.000000  0.000000  0.00000  0.00000000  0.0000000
#> cyl          0.00000  0.000000  0.000000  0.00000  0.00000000  0.0000000
#> disp         0.00000  0.000000  0.000000  0.00000  0.00000000  0.0000000
#> drat         0.00000  0.000000  0.000000  0.00000  0.00000000  0.0000000
#> gear         0.00000  0.000000  0.000000  0.00000  0.00000000  0.0000000
#> hp           0.00000  0.000000  0.000000  0.00000  0.00000000  0.0000000
#> qsec         0.00000  0.000000  0.000000  0.00000  0.04003869  0.1317208
#> vs           0.00000  0.000000  0.000000  0.00000  0.00000000  0.0000000
#> wt          -5.49909 -5.750004 -5.984008 -6.14803 -6.13666908 -6.1106541
#>                 1.4702     1.3711     1.2787     1.1925     1.1121     1.0372
#> (Intercept) 35.6056174 34.1164249 32.7275990 31.4323753 30.2244453 29.0979258
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.2172248  0.2969661  0.3713332  0.4406881  0.5053688  0.5656902
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -6.0863921 -6.0637653 -6.0426635 -6.0229839 -6.0046306 -5.9875143
#>                 0.9673     0.9021     0.8413     0.7846     0.7317     0.6824
#> (Intercept) 28.0473299 27.0675407 26.1537860 25.3016152 24.5068780 23.7657034
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.6219462  0.6744107  0.7233392  0.7689701  0.8115257  0.8512132
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.9715515 -5.9566646 -5.9427810 -5.9298331 -5.9177579 -5.9064965
#>                 0.6364     0.5935     0.5535     0.5162     0.4814     0.4490
#> (Intercept) 23.0744814 22.4298455 22.3851991 22.3851853 22.3851850 22.3851850
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.8882258  0.9227439  0.9251346  0.9251353  0.9251354  0.9251354
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.8959940 -5.8861995 -5.8855211 -5.8855209 -5.8855209 -5.8855209
#>                 0.4187     0.3905     0.3642     0.3396     0.3167     0.2954
#> (Intercept) 22.3851850 22.3851850 21.1670840 19.6300584 18.1954778 16.8618515
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.2508782  0.5673028  0.8626305  1.1372003
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.9251354  0.9251354  0.9235434  0.9215716  0.9197330  0.9180172
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.8855209 -5.8855209 -5.7728747 -5.6307874 -5.4981730 -5.3748816
#>                 0.2755    0.2569    0.2396    0.2235    0.2084    0.1943
#> (Intercept) 15.8248353 15.488167 15.489793 15.491214 15.492438 15.493482
#> am           0.0000000  2.740839  2.740044  2.739445  2.738991  2.738646
#> carb        -0.1188391 -1.433479 -1.433534 -1.433557 -1.433559 -1.433549
#> cyl          0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         1.6288268  1.830322  1.829914  1.829614  1.829394  1.829233
#> gear         0.0000000  1.589020  1.589427  1.589684  1.589840  1.589929
#> hp           0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.8479830  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.0000000  2.596447  2.596120  2.595883  2.595713  2.595589
#> wt          -5.0980320 -1.773271 -1.773618 -1.773912 -1.774158 -1.774364
#>                0.1812    0.1690    0.1576    0.1470    0.1371    0.1279
#> (Intercept) 15.494364 15.495106 15.495726 15.496242 15.496671 15.497027
#> am           2.738382  2.738180  2.738023  2.737902  2.737808  2.737733
#> carb        -1.433533 -1.433514 -1.433494 -1.433474 -1.433456 -1.433440
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         1.829114  1.829027  1.828962  1.828914  1.828879  1.828852
#> gear         1.589974  1.589991  1.589991  1.589981  1.589966  1.589948
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           2.595500  2.595436  2.595389  2.595355  2.595330  2.595312
#> wt          -1.774535 -1.774677 -1.774795 -1.774892 -1.774972 -1.775038
#>                  0.1192      0.1112      0.1037      0.0967      0.0902
#> (Intercept)  2.79536708  2.79331702  2.81722156  2.81607118  2.81495573
#> am           2.69602370  2.69605275  2.48489145  2.48485802  2.48482561
#> carb        -1.10892996 -1.10876797 -0.37569260 -0.37547912 -0.37527212
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.01935258  0.01935682  0.03466916  0.03467426  0.03467921
#> drat         2.01113692  2.01102789  1.71273088  1.71262252  1.71251746
#> gear         2.11802407  2.11796880  2.23913334  2.23913750  2.23914153
#> hp           0.00000000  0.00000000 -0.02303178 -0.02303569 -0.02303949
#> qsec         0.60780073  0.60798211  0.71453701  0.71465205  0.71476360
#> vs           2.46526463  2.46514245  2.67180657  2.67180394  2.67180138
#> wt          -3.70470730 -3.70534000 -4.84711999 -4.84767168 -4.84820661
#>                  0.0841      0.0785      0.0732      0.0682      0.0636
#> (Intercept)  2.81387416  2.81282544  2.81180858  2.81082259  2.80986655
#> am           2.48479418  2.48476371  2.48473416  2.48470551  2.48467773
#> carb        -0.37507140 -0.37487679 -0.37468808 -0.37450510 -0.37432768
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.03468401  0.03468867  0.03469318  0.03469756  0.03470180
#> drat         1.71241559  1.71231681  1.71222103  1.71212816  1.71203810
#> gear         2.23914544  2.23914923  2.23915290  2.23915646  2.23915992
#> hp          -0.02304316 -0.02304673 -0.02305018 -0.02305354 -0.02305679
#> qsec         0.71487176  0.71497664  0.71507833  0.71517694  0.71527255
#> vs           2.67179890  2.67179650  2.67179417  2.67179191  2.67178972
#> wt          -4.84872530 -4.84922823 -4.84971589 -4.85018874 -4.85064722
#>                  0.0594      0.0554      0.0516      0.0481      0.0449
#> (Intercept)  2.80893954  2.80804069  2.80716914  2.80632405  2.80550464
#> am           2.48465079  2.48462468  2.48459935  2.48457480  2.48455099
#> carb        -0.37415565 -0.37398885 -0.37382711 -0.37367028 -0.37351821
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.03470592  0.03470991  0.03471377  0.03471753  0.03472116
#> drat         1.71195079  1.71186613  1.71178403  1.71170443  1.71162725
#> gear         2.23916327  2.23916652  2.23916967  2.23917272  2.23917568
#> hp          -0.02305994 -0.02306299 -0.02306595 -0.02306883 -0.02307161
#> qsec         0.71536525  0.71545514  0.71554230  0.71562681  0.71570876
#> vs           2.67178759  2.67178553  2.67178354  2.67178160  2.67177972
#> wt          -4.85109179 -4.85152285 -4.85194082 -4.85234609 -4.85273906
#>                  0.0419      0.0390      0.0364      0.0340      0.0317
#> (Intercept)  2.80471010  2.80393970  2.80319270  2.80246839  2.80176607
#> am           2.48452790  2.48450552  2.48448381  2.48446276  2.48444236
#> carb        -0.37337077 -0.37322780 -0.37308917 -0.37295475 -0.37282442
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.03472469  0.03472811  0.03473142  0.03473464  0.03473776
#> drat         1.71155241  1.71147985  1.71140949  1.71134127  1.71127511
#> gear         2.23917855  2.23918134  2.23918404  2.23918665  2.23918919
#> hp          -0.02307431 -0.02307693 -0.02307947 -0.02308193 -0.02308432
#> qsec         0.71578821  0.71586526  0.71593996  0.71601240  0.71608263
#> vs           2.67177790  2.67177614  2.67177442  2.67177276  2.67177115
#> wt          -4.85312009 -4.85348955 -4.85384779 -4.85419515 -4.85453196
#>                  0.0295      0.0275      0.0257      0.0240      0.0223
#> (Intercept)  2.80108508  2.80042477  2.79978452  2.79916372  2.79856176
#> am           2.48442257  2.48440338  2.48438478  2.48436674  2.48434925
#> carb        -0.37269805 -0.37257551 -0.37245669 -0.37234148 -0.37222978
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.03474078  0.03474371  0.03474655  0.03474931  0.03475198
#> drat         1.71121097  1.71114878  1.71108847  1.71103000  1.71097330
#> gear         2.23919165  2.23919404  2.23919635  2.23919860  2.23920077
#> hp          -0.02308664 -0.02308888 -0.02309106 -0.02309317 -0.02309521
#> qsec         0.71615073  0.71621677  0.71628080  0.71634288  0.71640308
#> vs           2.67176959  2.67176808  2.67176661  2.67176519  2.67176381
#> wt          -4.85485854 -4.85517520 -4.85548225 -4.85577996 -4.85606864
#>                  0.0208      0.0194      0.0181      0.0169      0.0158
#> (Intercept)  2.79797809  2.79741215  2.79686339  2.79633130  2.79581537
#> am           2.48433229  2.48431584  2.48429990  2.48428444  2.48426945
#> carb        -0.37212146 -0.37201643 -0.37191460 -0.37181585 -0.37172011
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.03475457  0.03475708  0.03475952  0.03476188  0.03476417
#> drat         1.71091832  1.71086501  1.71081333  1.71076321  1.71071461
#> gear         2.23920288  2.23920493  2.23920691  2.23920883  2.23921070
#> hp          -0.02309720 -0.02309912 -0.02310099 -0.02310280 -0.02310455
#> qsec         0.71646145  0.71651805  0.71657292  0.71662614  0.71667773
#> vs           2.67176247  2.67176118  2.67175992  2.67175870  2.67175752
#> wt          -4.85634855 -4.85661996 -4.85688313 -4.85713830 -4.85738572
#>                  0.0147      0.0137      0.0128      0.0119      0.0111
#> (Intercept)  2.79531511  2.79483005  2.79435971  2.79390366  2.79346146
#> am           2.48425491  2.48424082  2.48422715  2.48421390  2.48420105
#> carb        -0.37162727 -0.37153726 -0.37144997 -0.37136534 -0.37128328
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.03476639  0.03476854  0.03477063  0.03477265  0.03477462
#> drat         1.71066749  1.71062180  1.71057750  1.71053455  1.71049290
#> gear         2.23921251  2.23921426  2.23921596  2.23921761  2.23921920
#> hp          -0.02310625 -0.02310790 -0.02310950 -0.02311105 -0.02311255
#> qsec         0.71672776  0.71677627  0.71682330  0.71686891  0.71691313
#> vs           2.67175637  2.67175526  2.67175418  2.67175314  2.67175212
#> wt          -4.85762563 -4.85785826 -4.85808381 -4.85830252 -4.85851459
#>                  0.0104      0.0097      0.0090      0.0084      0.0078
#> (Intercept)  2.79303269  2.79261694  2.79221382  2.79182294  2.79144393
#> am           2.48418859  2.48417651  2.48416480  2.48415344  2.48414243
#> carb        -0.37120371 -0.37112655 -0.37105174 -0.37097921 -0.37090887
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.03477652  0.03477837  0.03478016  0.03478189  0.03478357
#> drat         1.71045251  1.71041335  1.71037538  1.71033856  1.71030286
#> gear         2.23922075  2.23922226  2.23922371  2.23922512  2.23922649
#> hp          -0.02311401 -0.02311542 -0.02311679 -0.02311812 -0.02311941
#> qsec         0.71695601  0.71699759  0.71703790  0.71707699  0.71711490
#> vs           2.67175114  2.67175019  2.67174926  2.67174837  2.67174750
#> wt          -4.85872021 -4.85891959 -4.85911291 -4.85930037 -4.85948213
#>                  0.0073      0.0068      0.0064      0.0059      0.0055
#> (Intercept)  2.79107644  2.79072010  2.79037459  2.79003957  2.78971473
#> am           2.48413175  2.48412139  2.48411135  2.48410162  2.48409218
#> carb        -0.37084067 -0.37077455 -0.37071043 -0.37064825 -0.37058797
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.03478520  0.03478679  0.03478832  0.03478981  0.03479125
#> drat         1.71026825  1.71023468  1.71020214  1.71017058  1.71013999
#> gear         2.23922782  2.23922911  2.23923036  2.23923157  2.23923274
#> hp          -0.02312066 -0.02312187 -0.02312304 -0.02312418 -0.02312529
#> qsec         0.71715165  0.71718728  0.71722184  0.71725534  0.71728783
#> vs           2.67174666  2.67174584  2.67174505  2.67174428  2.67174354
#> wt          -4.85965837 -4.85982925 -4.85999495 -4.86015562 -4.86031140
#> 
#> $iter
#>   [1]   0   8   4   3   3   2   2   2   2   2   2   2   2   2   2   2   2   4
#>  [19]   3   3   3   3   3   3   3   3   3   3   3   3   3   3   3   2   1   1
#>  [37]   1   1   1  19  19  19  18  61 153   1   1   1   1   1   1   1   1   1
#>  [55]   1  96   1 164   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.535041084 5.161997840 4.814096462 4.489642473 4.187055597 3.904862063
#>   [7] 3.641687429 3.396249885 3.167354009 2.953884950 2.754802990 2.569138489
#>  [13] 2.395987153 2.234505638 2.083907437 1.943459050 1.812476415 1.690321571
#>  [19] 1.576399554 1.470155500 1.371071940 1.278666280 1.192488452 1.112118722
#>  [25] 1.037165642 0.967264149 0.902073783 0.841277029 0.784577773 0.731699857
#>  [31] 0.682385735 0.636395221 0.593504313 0.553504108 0.516199784 0.481409646
#>  [37] 0.448964247 0.418705560 0.390486206 0.364168743 0.339624988 0.316735401
#>  [43] 0.295388495 0.275480299 0.256913849 0.239598715 0.223450564 0.208390744
#>  [49] 0.194345905 0.181247641 0.169032157 0.157639955 0.147015550 0.137107194
#>  [55] 0.127866628 0.119248845 0.111211872 0.103716564 0.096726415 0.090207378
#>  [61] 0.084127703 0.078457777 0.073169986 0.068238574 0.063639522 0.059350431
#>  [67] 0.055350411 0.051619978 0.048140965 0.044896425 0.041870556 0.039048621
#>  [73] 0.036416874 0.033962499 0.031673540 0.029538849 0.027548030 0.025691385
#>  [79] 0.023959872 0.022345056 0.020839074 0.019434591 0.018124764 0.016903216
#>  [85] 0.015763996 0.014701555 0.013710719 0.012786663 0.011924885 0.011121187
#>  [91] 0.010371656 0.009672641 0.009020738 0.008412770 0.007845778 0.007316999
#>  [97] 0.006823857 0.006363952 0.005935043 0.005535041
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 39
#> 
#> $loss
#>   [1] 790.33143 666.82379 563.49128 477.43055 406.13498 347.44174 299.48588
#>   [8] 260.66042 229.58174 205.05954 186.07076 171.73701 161.30492 154.12917
#>  [15] 149.65772 147.41908 146.96115 141.47894 129.83899 120.12935 112.07077
#>  [22] 105.42218  99.97563  95.55191  91.99667  89.17709  86.97898  85.30426
#>  [29]  84.06877  83.20041  82.63746  82.32718  82.22457  82.22413  82.22413
#>  [36]  82.22413  82.22413  82.22413  82.22413  80.15828  78.10856  76.75495
#>  [43]  75.98104  73.51031  48.52348  48.52347  48.52347  48.52346  48.52346
#>  [50]  48.52346  48.52346  48.52346  48.52346  48.52346  48.52346  41.19864
#>  [57]  41.19863  37.52191  37.52189  37.52188  37.52186  37.52184  37.52183
#>  [64]  37.52182  37.52180  37.52179  37.52178  37.52177  37.52176  37.52175
#>  [71]  37.52174  37.52174  37.52173  37.52172  37.52171  37.52171  37.52170
#>  [78]  37.52170  37.52169  37.52169  37.52168  37.52168  37.52167  37.52167
#>  [85]  37.52167  37.52166  37.52166  37.52166  37.52165  37.52165  37.52165
#>  [92]  37.52164  37.52164  37.52164  37.52164  37.52164  37.52164  37.52163
#>  [99]  37.52163  37.52163
#> 
#> $linear.predictors
#>      5.5350   5.1620   4.8141   4.4896   4.1871   3.9049   3.6417   3.3962
#> 1  20.54286 20.84732 21.13126 21.39607 21.64303 21.87335 22.08814 22.28846
#> 2  20.54286 20.47751 20.41656 20.35972 20.30672 20.25728 20.21118 20.16818
#> 3  20.54286 20.33766 20.14630 19.96783 19.80139 19.64616 19.50140 19.36640
#> 4  20.54286 20.32523 20.12227 19.93299 19.75647 19.59184 19.43831 19.29513
#> 5  20.54286 20.25686 19.99014 19.74140 19.50942 19.29307 19.09131 18.90315
#> 6  20.54286 20.49305 20.44659 20.40327 20.36286 20.32518 20.29004 20.25727
#> 7  20.54286 20.51791 20.49464 20.47294 20.45270 20.43383 20.41622 20.39981
#> 8  20.54286 20.33766 20.14630 19.96783 19.80139 19.64616 19.50140 19.36640
#> 9  20.54286 20.12634 19.73789 19.37563 19.03778 18.72270 18.42885 18.15481
#> 10 20.54286 19.10454 17.76315 16.51218 15.34551 14.25747 13.24277 12.29645
#> 11 20.54286 21.47196 22.33845 23.14654 23.90017 24.60300 25.25847 25.86976
#> 12 20.54286 21.33523 22.07419 22.76335 23.40607 24.00546 24.56446 25.08579
#> 13 20.54286 20.28794 20.05020 19.82849 19.62171 19.42888 19.24904 19.08132
#> 14 20.54286 20.34077 20.15230 19.97654 19.81262 19.65974 19.51718 19.38422
#> 15 20.54286 20.08905 19.66582 19.27112 18.90302 18.55973 18.23958 17.94100
#> 16 20.54286 20.08594 19.65982 19.26241 18.89180 18.54615 18.22381 17.92319
#> 17 20.54286 21.27307 21.95407 22.58918 23.18148 23.73386 24.24901 24.72944
#> 18 20.54286 21.14566 21.70783 22.23212 22.72107 23.17706 23.60233 23.99893
#> 19 20.54286 21.53536 22.46097 23.32420 24.12925 24.88004 25.58023 26.23323
#> 20 20.54286 20.75409 20.95109 21.13481 21.30614 21.46593 21.61496 21.75393
#> 21 20.54286 20.25686 19.99014 19.74140 19.50942 19.29307 19.09131 18.90315
#>      3.1674   2.9539    2.7548    2.5691    2.3960    2.2345    2.0839
#> 1  22.47527 22.64950 22.811979 22.963511 23.104831 23.236625 23.359538
#> 2  20.12808 20.09069 20.055813 20.023288 19.992956 19.964667 19.938285
#> 3  19.24049 19.12307 19.013566 18.911439 18.816196 18.727372 18.644534
#> 4  19.16160 19.03706 18.920921 18.812608 18.711595 18.617390 18.529534
#> 5  18.72766 18.56400 18.411378 18.269038 18.136291 18.012490 17.897034
#> 6  20.22671 20.19820 20.171618 20.146827 20.123707 20.102144 20.082035
#> 7  20.38450 20.37022 20.356907 20.344489 20.332908 20.322108 20.312036
#> 8  19.24049 19.12307 19.013566 18.911439 18.816196 18.727372 18.644534
#> 9  17.89924 17.66090 17.438614 17.231312 17.037982 16.857681 16.689533
#> 10 11.41391 10.59085  9.823258  9.107402  8.439793  7.817178  7.236526
#> 11 26.43985 26.97152 27.467352 27.929770 28.361023 28.763211 29.138292
#> 12 25.57198 26.02540 26.448265 26.842629 27.210414 27.553411 27.873291
#> 13 18.92490 18.77903 18.642989 18.516115 18.397793 18.287445 18.184534
#> 14 19.26022 19.14457 19.036727 18.936147 18.842346 18.754867 18.673284
#> 15 17.66255 17.40287 17.160681 16.934819 16.724179 16.527736 16.344533
#> 16 17.64283 17.38136 17.137520 16.910111 16.698029 16.500241 16.315782
#> 17 25.17749 25.59535 25.985044 26.348474 26.687409 27.003502 27.298291
#> 18 24.36880 24.71374 25.035441 25.335456 25.615251 25.876189 26.119540
#> 19 26.84222 27.41017 27.939837 28.433808 28.894487 29.324118 29.724793
#> 20 21.88354 22.00442 22.117147 22.222279 22.320324 22.411762 22.497037
#> 21 18.72766 18.56400 18.411378 18.269038 18.136291 18.012490 17.897034
#>       1.9435    1.8125    1.6903    1.5764    1.4702    1.3711   1.2787
#> 1  23.474166 23.554514 23.495582 23.360640 23.234790 23.117422 23.00796
#> 2  19.913681 19.896436 19.963580 20.117328 20.260717 20.394441 20.51915
#> 3  18.567280 18.513129 18.485935 18.423667 18.365595 18.311436 18.26093
#> 4  18.447599 18.390168 18.491326 18.722960 18.938986 19.140452 19.32834
#> 5  17.789358 17.713885 17.640923 17.473851 17.318038 17.172726 17.03721
#> 6  20.063282 20.050136 20.139418 20.343858 20.534522 20.712336 20.87817
#> 7  20.302642 20.296058 20.500997 20.970275 21.407930 21.816089 22.19674
#> 8  18.567280 18.513129 18.537185 18.592270 18.643642 18.691553 18.73623
#> 9  16.532717 16.422799 16.438706 16.475131 16.509102 16.540783 16.57033
#> 10  6.695007  6.315437  6.342815  6.405506  6.463972  6.518498  6.56935
#> 11 29.488095 29.733284 29.745415 29.773192 29.799097 29.823257 29.84579
#> 12 28.171613 28.380717 28.450601 28.610623 28.759861 28.899042 29.02884
#> 13 18.088559 18.021286 17.988996 17.915057 17.846099 17.781790 17.72181
#> 14 18.597200 18.543869 18.527830 18.491102 18.456849 18.424905 18.39511
#> 15 16.173676 16.053917 15.966806 15.767335 15.581306 15.407814 15.24602
#> 16 16.143756 16.023177 16.001786 15.952804 15.907122 15.864520 15.82479
#> 17 27.573212 27.765914 27.796895 27.867836 27.933997 27.995699 28.05324
#> 18 26.346490 26.505568 26.450793 26.325367 26.208392 26.099302 25.99756
#> 19 30.098464 30.360383 30.306492 30.183091 30.068005 29.960676 29.86058
#> 20 22.576565 22.632309 22.536645 22.317590 22.113296 21.922770 21.74509
#> 21 17.789358 17.713885 17.591275 17.310518 17.048680 16.804488 16.57675
#>       1.1925    1.1121    1.0372    0.9673   0.9021    0.8413    0.7846
#> 1  22.905884 22.810684 22.721900 22.639099 22.56188 22.489863 22.422701
#> 2  20.635460 20.743928 20.845085 20.939425 21.02741 21.109459 21.185981
#> 3  18.213823 18.169893 18.128924 18.090717 18.05508 18.021853 17.990861
#> 4  19.503565 19.666981 19.819383 19.961513 20.09406 20.217682 20.332969
#> 5  16.910823 16.792956 16.683033 16.580518 16.48491 16.395751 16.312598
#> 6  21.032820 21.177050 21.311560 21.437004 21.55399 21.663099 21.764850
#> 7  22.551735 22.882805 23.191562 23.479510 23.74805 23.998494 24.232057
#> 8  18.777904 18.816765 18.853008 18.886808 18.91833 18.947727 18.975143
#> 9  16.597883 16.623580 16.647546 16.669896 16.69074 16.710179 16.728309
#> 10  6.616774  6.661001  6.702248  6.740715  6.77659  6.810046  6.841248
#> 11 29.866801 29.886397 29.904673 29.921717 29.93761 29.952437 29.966262
#> 12 29.149894 29.262788 29.368073 29.466262 29.55783 29.643233 29.722877
#> 13 17.665881 17.613718 17.565070 17.519700 17.47739 17.437929 17.401129
#> 14 18.367331 18.341420 18.317255 18.294719 18.27370 18.254101 18.235822
#> 15 15.095122 14.954397 14.823157 14.700763 14.58662 14.480164 14.380886
#> 16 15.787735 15.753179 15.720952 15.690897 15.66287 15.636726 15.612348
#> 17 28.106908 28.156956 28.203631 28.247160 28.28776 28.325616 28.360924
#> 18 25.902682 25.814195 25.731672 25.654711 25.58294 25.515999 25.453574
#> 19 29.767230 29.680172 29.598981 29.523263 29.45265 29.386791 29.325373
#> 20 21.579376 21.424835 21.280710 21.146298 21.02094 20.904040 20.795015
#> 21 16.364370 16.166299 15.981577 15.809305 15.64864 15.498810 15.359075
#>       0.7317    0.6824    0.6364    0.5935    0.5535    0.5162    0.4814
#> 1  22.360065 22.301651 22.247174 22.196368 22.192849 22.192848 22.192848
#> 2  21.257346 21.323901 21.385970 21.443856 21.447865 21.447867 21.447867
#> 3  17.961958 17.935003 17.909865 17.886421 17.884798 17.884797 17.884797
#> 4  20.440485 20.540756 20.634268 20.721478 20.727518 20.727520 20.727520
#> 5  16.235049 16.162727 16.095279 16.032377 16.028021 16.028020 16.028020
#> 6  21.859744 21.948243 22.030777 22.107748 22.113079 22.113081 22.113081
#> 7  24.449879 24.653021 24.842471 25.019153 25.031390 25.031394 25.031394
#> 8  19.000711 19.024556 19.046794 19.067533 19.068970 19.068970 19.068970
#> 9  16.745216 16.760983 16.775688 16.789402 16.790352 16.790353 16.790353
#> 10  6.870347  6.897485  6.922794  6.946397  6.948031  6.948032  6.948032
#> 11 29.979155 29.991179 30.002393 30.012851 30.013575 30.013576 30.013576
#> 12 29.797154 29.866424 29.931026 29.991274 29.995447 29.995448 29.995448
#> 13 17.366809 17.334802 17.304952 17.277114 17.275186 17.275185 17.275185
#> 14 18.218774 18.202876 18.188048 18.174220 18.173263 18.173263 18.173262
#> 15 14.288299 14.201952 14.121424 14.046324 14.041122 14.041121 14.041121
#> 16 15.589612 15.568409 15.548634 15.530193 15.528916 15.528915 15.528915
#> 17 28.393852 28.424561 28.453201 28.479910 28.481760 28.481760 28.481760
#> 18 25.395355 25.341061 25.290425 25.243202 25.239932 25.239931 25.239931
#> 19 29.268095 29.214676 29.164859 29.118398 29.115181 29.115180 29.115179
#> 20 20.693337 20.598512 20.510078 20.427604 20.421892 20.421890 20.421890
#> 21 15.228758 15.107223 14.993879 14.888175 14.880854 14.880852 14.880852
#>       0.4490    0.4187    0.3905   0.3642    0.3396    0.3167    0.2954
#> 1  22.192848 22.192848 22.192848 22.22210 22.258945 22.293329 22.325306
#> 2  21.447867 21.447867 21.447867 21.33368 21.189721 21.055363 20.930438
#> 3  17.884797 17.884797 17.884797 17.81737 17.732302 17.652904 17.579092
#> 4  20.727520 20.727520 20.727520 20.55941 20.347467 20.149661 19.965741
#> 5  16.028020 16.028020 16.028020 15.99217 15.946883 15.904615 15.865329
#> 6  22.113081 22.113081 22.113081 22.14822 22.192626 22.234073 22.272592
#> 7  25.031394 25.031394 25.031394 25.11511 25.220894 25.319630 25.411393
#> 8  19.068970 19.068970 19.068970 19.19268 19.348737 19.494388 19.629798
#> 9  16.790353 16.790353 16.790353 16.73959 16.675590 16.615854 16.560313
#> 10  6.948032  6.948032  6.948032  7.06519  7.212982  7.350921  7.479161
#> 11 30.013576 30.013576 30.013576 30.18474 30.400645 30.602152 30.789493
#> 12 29.995448 29.995448 29.995448 30.01108 30.030856 30.049318 30.066471
#> 13 17.275185 17.275185 17.275185 17.11917 16.922355 16.738665 16.567891
#> 14 18.173262 18.173262 18.173262 18.10483 18.018496 17.937921 17.863011
#> 15 14.041121 14.041121 14.041121 14.16682 14.325292 14.473191 14.610708
#> 16 15.528915 15.528915 15.528915 15.48950 15.439769 15.393352 15.350201
#> 17 28.481760 28.481760 28.481760 28.47512 28.466783 28.458999 28.451757
#> 18 25.239931 25.239931 25.239931 25.34770 25.483570 25.610382 25.728289
#> 19 29.115179 29.115179 29.115179 28.98642 28.823968 28.672347 28.531391
#> 20 20.421890 20.421890 20.421890 20.39932 20.370773 20.344123 20.319361
#> 21 14.880852 14.880852 14.880852 14.92976 14.991344 15.048814 15.102264
#>       0.2755   0.2569   0.2396   0.2235   0.2084   0.1943   0.1812   0.1690
#> 1  22.302860 21.34346 21.34319 21.34301 21.34290 21.34282 21.34277 21.34274
#> 2  20.817399 21.35452 21.35461 21.35468 21.35473 21.35477 21.35479 21.35481
#> 3  17.613402 17.05373 17.05399 17.05418 17.05433 17.05444 17.05453 17.05460
#> 4  19.708583 20.33437 20.33450 20.33459 20.33465 20.33469 20.33472 20.33473
#> 5  15.810089 14.06607 14.06614 14.06624 14.06633 14.06643 14.06652 14.06660
#> 6  22.294466 22.67089 22.67109 22.67122 22.67129 22.67134 22.67136 22.67137
#> 7  25.332168 23.16280 23.16292 23.16298 23.16302 23.16303 23.16304 23.16303
#> 8  19.715339 19.78159 19.78150 19.78144 19.78140 19.78137 19.78135 19.78135
#> 9  16.461949 14.87092 14.87103 14.87112 14.87120 14.87127 14.87132 14.87137
#> 10  7.695291 10.39406 10.39357 10.39318 10.39287 10.39262 10.39242 10.39225
#> 11 31.088596 30.47423 30.47368 30.47329 30.47303 30.47285 30.47273 30.47265
#> 12 30.099618 30.21806 30.21778 30.21756 30.21741 30.21729 30.21720 30.21713
#> 13 16.443119 16.19804 16.19843 16.19872 16.19893 16.19909 16.19921 16.19930
#> 14 17.876327 17.06260 17.06285 17.06305 17.06320 17.06331 17.06340 17.06347
#> 15 14.915978 14.53905 14.53882 14.53868 14.53860 14.53855 14.53853 14.53853
#> 16 15.460121 16.20743 16.20758 16.20767 16.20774 16.20778 16.20780 16.20782
#> 17 28.513796 29.78449 29.78423 29.78403 29.78388 29.78376 29.78367 29.78360
#> 18 26.054387 27.62068 27.62088 27.62098 27.62103 27.62104 27.62102 27.62099
#> 19 28.345424 30.12095 30.12131 30.12156 30.12174 30.12186 30.12194 30.12199
#> 20 20.030341 19.28704 19.28713 19.28721 19.28726 19.28731 19.28735 19.28738
#> 21 14.820747 14.85504 14.85477 14.85459 14.85447 14.85438 14.85433 14.85429
#>      0.1576   0.1470   0.1371   0.1279   0.1192   0.1112   0.1037   0.0967
#> 1  21.34273 21.34272 21.34272 21.34272 20.76568 20.76567 21.51092 21.51102
#> 2  21.35483 21.35484 21.35485 21.35486 21.59805 21.59813 21.82420 21.82429
#> 3  17.05465 17.05469 17.05472 17.05474 17.83417 17.83437 18.11600 18.11616
#> 4  20.33474 20.33474 20.33474 20.33474 19.88229 19.88224 19.61700 19.61693
#> 5  14.06667 14.06674 14.06679 14.06684 14.53816 14.53838 14.38187 14.38198
#> 6  22.67136 22.67136 22.67135 22.67134 22.11299 22.11274 22.50058 22.50051
#> 7  23.16302 23.16301 23.16299 23.16298 24.37218 24.37244 24.19595 24.19606
#> 8  19.78134 19.78134 19.78134 19.78134 18.80272 18.80240 19.03627 19.03615
#> 9  14.87141 14.87144 14.87146 14.87148 14.27090 14.27088 13.62121 13.62107
#> 10 10.39212 10.39202 10.39193 10.39186 10.38599 10.38585 10.60829 10.60825
#> 11 30.47260 30.47257 30.47255 30.47254 30.86415 30.86423 31.45476 31.45491
#> 12 30.21708 30.21704 30.21701 30.21699 30.47988 30.47997 30.07522 30.07519
#> 13 16.19937 16.19942 16.19946 16.19949 15.84947 15.84945 16.07277 16.07280
#> 14 17.06352 17.06356 17.06359 17.06362 16.93913 16.93915 16.97463 16.97464
#> 15 14.53854 14.53855 14.53857 14.53858 14.12880 14.12867 13.30983 13.30962
#> 16 16.20783 16.20783 16.20783 16.20783 16.98532 16.98544 17.44123 17.44138
#> 17 29.78354 29.78350 29.78347 29.78344 29.37294 29.37283 28.88704 28.88689
#> 18 27.62096 27.62092 27.62089 27.62085 27.32330 27.32303 26.96853 26.96831
#> 19 30.12203 30.12205 30.12206 30.12207 30.41794 30.41795 30.31162 30.31163
#> 20 19.28740 19.28742 19.28744 19.28745 18.67324 18.67320 19.08898 19.08901
#> 21 14.85426 14.85425 14.85424 14.85423 15.80271 15.80299 15.40310 15.40319
#>      0.0902   0.0841   0.0785   0.0732   0.0682   0.0636   0.0594   0.0554
#> 1  21.51112 21.51121 21.51130 21.51139 21.51147 21.51155 21.51163 21.51171
#> 2  21.82438 21.82446 21.82454 21.82462 21.82470 21.82477 21.82484 21.82491
#> 3  18.11632 18.11647 18.11662 18.11676 18.11690 18.11703 18.11716 18.11729
#> 4  19.61686 19.61679 19.61673 19.61667 19.61661 19.61655 19.61649 19.61644
#> 5  14.38207 14.38217 14.38226 14.38235 14.38244 14.38252 14.38260 14.38268
#> 6  22.50045 22.50040 22.50034 22.50029 22.50023 22.50018 22.50013 22.50008
#> 7  24.19617 24.19627 24.19637 24.19647 24.19656 24.19665 24.19674 24.19683
#> 8  19.03603 19.03591 19.03579 19.03568 19.03558 19.03547 19.03537 19.03527
#> 9  13.62094 13.62080 13.62068 13.62055 13.62043 13.62032 13.62020 13.62009
#> 10 10.60821 10.60817 10.60813 10.60809 10.60805 10.60802 10.60798 10.60795
#> 11 31.45505 31.45518 31.45532 31.45545 31.45557 31.45569 31.45581 31.45592
#> 12 30.07516 30.07514 30.07511 30.07509 30.07506 30.07504 30.07502 30.07500
#> 13 16.07283 16.07285 16.07287 16.07290 16.07292 16.07294 16.07296 16.07298
#> 14 16.97464 16.97465 16.97466 16.97467 16.97468 16.97468 16.97469 16.97470
#> 15 13.30943 13.30923 13.30905 13.30887 13.30869 13.30852 13.30836 13.30820
#> 16 17.44153 17.44167 17.44181 17.44195 17.44208 17.44221 17.44233 17.44245
#> 17 28.88675 28.88661 28.88648 28.88635 28.88623 28.88610 28.88599 28.88587
#> 18 26.96810 26.96790 26.96770 26.96751 26.96732 26.96714 26.96697 26.96680
#> 19 30.31164 30.31164 30.31165 30.31165 30.31166 30.31166 30.31167 30.31167
#> 20 19.08905 19.08908 19.08912 19.08915 19.08918 19.08921 19.08924 19.08927
#> 21 15.40329 15.40338 15.40347 15.40356 15.40364 15.40373 15.40380 15.40388
#>      0.0516   0.0481   0.0449   0.0419   0.0390   0.0364   0.0340   0.0317
#> 1  21.51179 21.51186 21.51193 21.51200 21.51207 21.51213 21.51219 21.51225
#> 2  21.82498 21.82505 21.82511 21.82517 21.82523 21.82529 21.82535 21.82540
#> 3  18.11741 18.11753 18.11765 18.11776 18.11787 18.11797 18.11808 18.11817
#> 4  19.61638 19.61633 19.61628 19.61623 19.61618 19.61614 19.61609 19.61605
#> 5  14.38276 14.38283 14.38290 14.38297 14.38304 14.38311 14.38317 14.38323
#> 6  22.50004 22.49999 22.49995 22.49990 22.49986 22.49982 22.49978 22.49975
#> 7  24.19691 24.19699 24.19707 24.19714 24.19722 24.19729 24.19736 24.19743
#> 8  19.03518 19.03509 19.03500 19.03491 19.03483 19.03475 19.03467 19.03459
#> 9  13.61999 13.61988 13.61978 13.61969 13.61959 13.61950 13.61941 13.61933
#> 10 10.60791 10.60788 10.60785 10.60782 10.60779 10.60777 10.60774 10.60771
#> 11 31.45603 31.45614 31.45624 31.45635 31.45644 31.45654 31.45663 31.45672
#> 12 30.07497 30.07495 30.07493 30.07492 30.07490 30.07488 30.07486 30.07484
#> 13 16.07300 16.07302 16.07304 16.07306 16.07307 16.07309 16.07311 16.07312
#> 14 16.97470 16.97471 16.97472 16.97472 16.97473 16.97473 16.97474 16.97474
#> 15 13.30804 13.30789 13.30774 13.30760 13.30747 13.30733 13.30720 13.30708
#> 16 17.44257 17.44268 17.44279 17.44290 17.44300 17.44310 17.44320 17.44329
#> 17 28.88576 28.88565 28.88555 28.88545 28.88535 28.88526 28.88516 28.88507
#> 18 26.96663 26.96647 26.96632 26.96617 26.96602 26.96588 26.96575 26.96561
#> 19 30.31168 30.31168 30.31169 30.31169 30.31169 30.31170 30.31170 30.31171
#> 20 19.08930 19.08932 19.08935 19.08938 19.08940 19.08942 19.08945 19.08947
#> 21 15.40396 15.40403 15.40410 15.40417 15.40423 15.40430 15.40436 15.40442
#>      0.0295   0.0275   0.0257   0.0240   0.0223   0.0208   0.0194   0.0181
#> 1  21.51231 21.51237 21.51242 21.51248 21.51253 21.51258 21.51263 21.51268
#> 2  21.82545 21.82550 21.82555 21.82560 21.82565 21.82569 21.82574 21.82578
#> 3  18.11827 18.11836 18.11845 18.11854 18.11862 18.11871 18.11879 18.11886
#> 4  19.61601 19.61597 19.61593 19.61589 19.61586 19.61582 19.61578 19.61575
#> 5  14.38329 14.38335 14.38341 14.38346 14.38351 14.38357 14.38362 14.38366
#> 6  22.49971 22.49968 22.49964 22.49961 22.49958 22.49954 22.49951 22.49948
#> 7  24.19749 24.19755 24.19762 24.19767 24.19773 24.19779 24.19784 24.19789
#> 8  19.03452 19.03445 19.03438 19.03431 19.03425 19.03418 19.03412 19.03406
#> 9  13.61924 13.61916 13.61909 13.61901 13.61894 13.61887 13.61880 13.61873
#> 10 10.60769 10.60766 10.60764 10.60761 10.60759 10.60757 10.60755 10.60753
#> 11 31.45680 31.45689 31.45697 31.45705 31.45712 31.45720 31.45727 31.45734
#> 12 30.07483 30.07481 30.07480 30.07478 30.07477 30.07475 30.07474 30.07473
#> 13 16.07314 16.07315 16.07317 16.07318 16.07320 16.07321 16.07322 16.07323
#> 14 16.97475 16.97475 16.97476 16.97476 16.97477 16.97477 16.97478 16.97478
#> 15 13.30696 13.30684 13.30673 13.30662 13.30651 13.30640 13.30630 13.30621
#> 16 17.44338 17.44347 17.44356 17.44364 17.44372 17.44380 17.44387 17.44395
#> 17 28.88499 28.88490 28.88482 28.88474 28.88467 28.88459 28.88452 28.88445
#> 18 26.96549 26.96536 26.96524 26.96512 26.96501 26.96490 26.96479 26.96469
#> 19 30.31171 30.31171 30.31172 30.31172 30.31172 30.31173 30.31173 30.31173
#> 20 19.08949 19.08951 19.08953 19.08955 19.08957 19.08959 19.08961 19.08963
#> 21 15.40448 15.40453 15.40459 15.40464 15.40469 15.40474 15.40479 15.40484
#>      0.0169   0.0158   0.0147   0.0137   0.0128   0.0119   0.0111   0.0104
#> 1  21.51272 21.51277 21.51281 21.51285 21.51289 21.51293 21.51297 21.51301
#> 2  21.82582 21.82586 21.82590 21.82594 21.82597 21.82601 21.82604 21.82608
#> 3  18.11894 18.11901 18.11908 18.11915 18.11922 18.11928 18.11934 18.11940
#> 4  19.61572 19.61569 19.61566 19.61563 19.61560 19.61557 19.61554 19.61552
#> 5  14.38371 14.38376 14.38380 14.38384 14.38388 14.38392 14.38396 14.38400
#> 6  22.49946 22.49943 22.49940 22.49938 22.49935 22.49933 22.49930 22.49928
#> 7  24.19795 24.19799 24.19804 24.19809 24.19813 24.19818 24.19822 24.19826
#> 8  19.03400 19.03395 19.03389 19.03384 19.03379 19.03374 19.03369 19.03365
#> 9  13.61866 13.61860 13.61854 13.61848 13.61842 13.61837 13.61831 13.61826
#> 10 10.60751 10.60749 10.60747 10.60745 10.60743 10.60742 10.60740 10.60738
#> 11 31.45741 31.45747 31.45754 31.45760 31.45766 31.45771 31.45777 31.45783
#> 12 30.07471 30.07470 30.07469 30.07468 30.07467 30.07466 30.07464 30.07463
#> 13 16.07325 16.07326 16.07327 16.07328 16.07329 16.07330 16.07331 16.07332
#> 14 16.97478 16.97479 16.97479 16.97480 16.97480 16.97480 16.97481 16.97481
#> 15 13.30611 13.30602 13.30593 13.30584 13.30576 13.30568 13.30560 13.30552
#> 16 17.44402 17.44409 17.44415 17.44422 17.44428 17.44434 17.44440 17.44446
#> 17 28.88438 28.88432 28.88425 28.88419 28.88413 28.88407 28.88402 28.88396
#> 18 26.96459 26.96449 26.96440 26.96431 26.96422 26.96413 26.96405 26.96397
#> 19 30.31173 30.31174 30.31174 30.31174 30.31174 30.31175 30.31175 30.31175
#> 20 19.08964 19.08966 19.08968 19.08969 19.08971 19.08972 19.08973 19.08975
#> 21 15.40488 15.40493 15.40497 15.40501 15.40505 15.40509 15.40513 15.40516
#>      0.0097   0.0090   0.0084   0.0078   0.0073   0.0068   0.0064   0.0059
#> 1  21.51304 21.51308 21.51311 21.51314 21.51317 21.51321 21.51323 21.51326
#> 2  21.82611 21.82614 21.82617 21.82620 21.82623 21.82626 21.82628 21.82631
#> 3  18.11946 18.11952 18.11957 18.11963 18.11968 18.11973 18.11978 18.11982
#> 4  19.61549 19.61547 19.61544 19.61542 19.61540 19.61537 19.61535 19.61533
#> 5  14.38404 14.38407 14.38411 14.38414 14.38417 14.38420 14.38423 14.38426
#> 6  22.49926 22.49924 22.49921 22.49919 22.49917 22.49916 22.49914 22.49912
#> 7  24.19830 24.19834 24.19838 24.19841 24.19845 24.19848 24.19851 24.19855
#> 8  19.03360 19.03356 19.03351 19.03347 19.03343 19.03339 19.03336 19.03332
#> 9  13.61821 13.61816 13.61811 13.61807 13.61802 13.61798 13.61794 13.61790
#> 10 10.60737 10.60735 10.60734 10.60732 10.60731 10.60730 10.60728 10.60727
#> 11 31.45788 31.45793 31.45798 31.45803 31.45807 31.45812 31.45816 31.45820
#> 12 30.07462 30.07461 30.07461 30.07460 30.07459 30.07458 30.07457 30.07456
#> 13 16.07333 16.07334 16.07335 16.07336 16.07337 16.07337 16.07338 16.07339
#> 14 16.97481 16.97482 16.97482 16.97482 16.97482 16.97483 16.97483 16.97483
#> 15 13.30545 13.30538 13.30531 13.30524 13.30518 13.30511 13.30505 13.30499
#> 16 17.44451 17.44457 17.44462 17.44467 17.44472 17.44477 17.44481 17.44486
#> 17 28.88391 28.88386 28.88381 28.88376 28.88371 28.88367 28.88363 28.88358
#> 18 26.96389 26.96381 26.96374 26.96367 26.96360 26.96353 26.96347 26.96340
#> 19 30.31175 30.31176 30.31176 30.31176 30.31176 30.31176 30.31177 30.31177
#> 20 19.08976 19.08977 19.08979 19.08980 19.08981 19.08982 19.08983 19.08984
#> 21 15.40520 15.40523 15.40527 15.40530 15.40533 15.40536 15.40539 15.40542
#>      0.0055
#> 1  21.51329
#> 2  21.82633
#> 3  18.11987
#> 4  19.61531
#> 5  14.38429
#> 6  22.49910
#> 7  24.19858
#> 8  19.03328
#> 9  13.61786
#> 10 10.60726
#> 11 31.45825
#> 12 30.07455
#> 13 16.07340
#> 14 16.97483
#> 15 13.30493
#> 16 17.44490
#> 17 28.88354
#> 18 26.96334
#> 19 30.31177
#> 20 19.08985
#> 21 15.40545
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.4 18.7 18.1 14.3 24.4 22.8 19.2 15.2 10.4 30.4 33.9 15.5 15.2 13.3
#> [16] 19.2 27.3 26.0 30.4 19.7 15.0
#> 
#> $X
#>               am       carb        cyl        disp        drat       gear
#>  [1,]  1.2747549  0.6860266 -0.1097643 -0.59484669  0.58708354  0.3735437
#>  [2,] -0.7844645 -1.0427605 -0.1097643  0.24762423 -0.88019927 -0.9338592
#>  [3,] -0.7844645 -0.4664981  1.0427605  1.12448171 -0.75494342 -0.9338592
#>  [4,] -0.7844645 -1.0427605 -0.1097643 -0.03606496 -1.45279744 -0.9338592
#>  [5,] -0.7844645  0.6860266  1.0427605  1.12448171 -0.64758126 -0.9338592
#>  [6,] -0.7844645 -0.4664981 -1.2622890 -0.70918203  0.21131599  0.3735437
#>  [7,] -0.7844645 -0.4664981 -1.2622890 -0.75990222  0.62287092  0.3735437
#>  [8,] -0.7844645  0.6860266 -0.1097643 -0.52951221  0.62287092  0.3735437
#>  [9,] -0.7844645  0.1097643  1.0427605  0.40064445 -0.89809296 -0.9338592
#> [10,] -0.7844645  0.6860266  1.0427605  1.98414591 -1.02334881 -0.9338592
#> [11,]  1.2747549 -0.4664981 -1.2622890 -1.31954361  2.43013389  0.3735437
#> [12,]  1.2747549 -1.0427605 -1.2622890 -1.35908816  1.15968171  0.3735437
#> [13,] -0.7844645 -0.4664981  1.0427605  0.76342274 -1.45279744 -0.9338592
#> [14,] -0.7844645 -0.4664981  1.0427605  0.64306976 -0.75494342 -0.9338592
#> [15,] -0.7844645  0.6860266  1.0427605  1.03851529  0.28289076 -0.9338592
#> [16,] -0.7844645 -0.4664981  1.0427605  1.46834739 -0.88019927 -0.9338592
#> [17,]  1.2747549 -1.0427605 -1.2622890 -1.29117469  0.90917001  0.3735437
#> [18,]  1.2747549 -0.4664981 -1.2622890 -0.93613338  1.53544925  1.6809466
#> [19,]  1.2747549 -0.4664981 -1.2622890 -1.15276875  0.35446553  1.6809466
#> [20,]  1.2747549  1.8385514 -0.1097643 -0.72379632  0.08606014  1.6809466
#> [21,]  1.2747549  2.9910761  1.0427605  0.61727983 -0.05708940  1.6809466
#>                hp        qsec         vs          wt
#>  [1,] -0.49164016 -0.69602204 -0.8660254 -0.54410734
#>  [2,] -0.49164016  0.86014299  1.1547005  0.11678711
#>  [3,]  0.43145313 -0.40358834 -0.8660254  0.36670518
#>  [4,] -0.56264734  1.26746136  1.1547005  0.38892012
#>  [5,]  1.42555360 -1.01978792 -0.8660254  0.51110229
#>  [6,] -1.17330906  1.15257669  1.1547005  0.08901844
#>  [7,] -0.70466169  2.66696548  1.1547005  0.04458856
#>  [8,] -0.30702151  0.26483154  1.1547005  0.36670518
#>  [9,]  0.50246031  0.10817063 -0.8660254  0.74435915
#> [10,]  0.99951054  0.01417408 -0.8660254  2.57042718
#> [11,] -1.31532341  0.37971620  1.1547005 -1.66040805
#> [12,] -1.13070475  1.10035639  1.1547005 -1.41604372
#> [13,]  0.07641725 -0.48191880 -0.8660254  0.45556494
#> [14,]  0.07641725 -0.25737149 -0.8660254  0.36115145
#> [15,]  1.42555360 -1.24433522 -0.8660254  0.81100397
#> [16,]  0.43145313 -0.38792225 -0.8660254  0.81655771
#> [17,] -1.11650332  0.57815336  1.1547005 -1.30496902
#> [18,] -0.76146744 -0.57069331 -0.8660254 -1.07726589
#> [19,] -0.44903586 -0.46625271  1.1547005 -1.77370424
#> [20,]  0.43145313 -1.19733695 -0.8660254 -0.37749529
#> [21,]  2.70368278 -1.66731968 -0.8660254  0.51110229
#> attr(,"center")
#>  [1]   0.3809524   2.8095238   6.1904762 229.1952381   3.5719048   3.7142857
#>  [7] 144.6190476  17.7928571   0.4285714   3.1098571
#> attr(,"scale")
#>  [1]   0.4856209   1.7353207   1.7353207 116.3244906   0.5588561   0.7648752
#>  [7]  70.4154178   1.9149640   0.4948717   0.9002950
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 34
#> 
#> $lambda.min
#> [1] 0.5535041
#> 
#> $null.dev
#> [1] 37.63483
#> 
#> $Bias
#> [1] 2.849962
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 12.22916