Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 34.11266 32.33043 29.27386 26.23718 23.75377 21.74093 20.12746 18.85209
#>   [9] 17.86216 17.11246 16.56418 16.18411 15.94380 15.81898 15.78892 15.86535
#>  [17] 15.84436 15.63079 15.29579 14.94019 14.58461 14.20772 13.89435 13.26000
#>  [25] 12.55058 11.94566 11.43096 11.29852 11.23293 11.31366 11.90368 12.02419
#>  [33] 12.11553 12.88825 13.99531 14.20015 14.24129 14.39433 15.73655 15.85573
#>  [41] 15.41942 15.74810 15.96792 17.55330 17.68698 31.89516 31.56193 31.57106
#>  [49] 31.57243 31.78438 31.78580 32.72736 32.72873 32.68501 35.22100 34.10075
#>  [57] 39.33765 39.69087 39.68949 40.03476 43.40097 43.75254 43.94363 44.01549
#>  [65] 44.01759 43.84221 44.37314 44.37531 44.57618 44.48860 44.31240 44.31408
#>  [73] 44.31572 44.26710 44.02568 44.02635 44.02703 44.17739 44.23332 44.02021
#>  [81] 44.01827 43.95402 43.95180 43.96491 43.96290 43.97845 43.98331 43.98148
#>  [89] 44.27790 44.27909 44.28026 44.28141 44.28254 44.28365 44.28475 44.28586
#>  [97] 44.30046 44.30175 44.30302 44.30427
#> 
#> $cvse
#>   [1] 11.350705 11.354190 10.852009 10.051466  9.373189  8.788668  8.276917
#>   [8]  7.823078  7.417069  7.052386  6.725089  6.432985  6.174965  5.950496
#>  [15]  5.759220  5.598604  5.451399  5.303729  5.187121  5.120183  5.054766
#>  [22]  4.990319  4.935595  4.684875  4.376881  4.112097  3.885079  3.842798
#>  [29]  3.814474  3.787056  3.744949  3.726562  3.708132  4.253042  4.686135
#>  [36]  4.708575  4.721424  4.743537  5.420284  5.417759  5.404318  5.395353
#>  [43]  5.397875  5.560861  5.560038 16.830946 16.826238 16.826838 16.826617
#>  [50] 16.810126 16.811240 16.799187 16.800238 16.803390 16.849631 16.775481
#>  [57] 17.553199 17.526737 17.525741 17.504178 19.017149 18.986563 19.011122
#>  [64] 19.005598 19.007125 19.013824 18.995533 18.996931 18.978398 18.979369
#>  [71] 18.963221 18.964445 18.965635 18.980433 18.995538 18.996429 18.997300
#>  [78] 18.998986 19.064330 19.071572 19.069925 19.073249 19.071699 19.069207
#>  [85] 19.067720 19.066843 19.066750 19.065410 19.219244 19.220804 19.222337
#>  [92] 19.223842 19.225320 19.226772 19.228197 19.229591 19.229724 19.231049
#>  [99] 19.232350 19.233627
#> 
#> $fold
#>  [1]  5  7  3  2  4  8  5 10  3  8  1 10  6  2  7  9  6  4  9  1  1
#> 
#> $lambda
#>   [1] 4.789467731 4.466673636 4.165634782 3.884884940 3.623056697 3.378874802
#>   [7] 3.151149949 2.938772989 2.740709525 2.555994876 2.383729376 2.223073994
#>  [13] 2.073246247 1.933516389 1.803203856 1.681673951 1.568334755 1.462634242
#>  [19] 1.364057589 1.272124673 1.186387727 1.106429165 1.031859543 0.962315663
#>  [25] 0.897458808 0.836973088 0.780563903 0.727956509 0.678894676 0.633139447
#>  [31] 0.590467968 0.550672403 0.513558926 0.478946773 0.446667364 0.416563478
#>  [37] 0.388488494 0.362305670 0.337887480 0.315114995 0.293877299 0.274070953
#>  [43] 0.255599488 0.238372938 0.222307399 0.207324625 0.193351639 0.180320386
#>  [49] 0.168167395 0.156833475 0.146263424 0.136405759 0.127212467 0.118638773
#>  [55] 0.110642917 0.103185954 0.096231566 0.089745881 0.083697309 0.078056390
#>  [61] 0.072795651 0.067889468 0.063313945 0.059046797 0.055067240 0.051355893
#>  [67] 0.047894677 0.044666736 0.041656348 0.038848849 0.036230567 0.033788748
#>  [73] 0.031511499 0.029387730 0.027407095 0.025559949 0.023837294 0.022230740
#>  [79] 0.020732462 0.019335164 0.018032039 0.016816740 0.015683348 0.014626342
#>  [85] 0.013640576 0.012721247 0.011863877 0.011064292 0.010318595 0.009623157
#>  [91] 0.008974588 0.008369731 0.007805639 0.007279565 0.006788947 0.006331394
#>  [97] 0.005904680 0.005506724 0.005135589 0.004789468
#> 
#> $fit
#> $beta
#>               4.7895     4.4667    4.1656    3.8849    3.6231    3.3789
#> (Intercept) 20.05714 21.6813119 23.196017 24.608637 25.926051 27.154675
#> am           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000 -0.4999494 -0.966204 -1.401034 -1.806559 -2.184752
#>                3.1511   2.9388    2.7407    2.5560    2.3837    2.2231
#> (Intercept) 28.300494 29.36909 30.365664 31.295073 32.161843 32.970196
#> am           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.537457 -2.86639 -3.173154 -3.459244 -3.726052 -3.974878
#>                2.0732   1.9335   1.8032    1.6817    1.5683     1.4626
#> (Intercept) 33.724069 34.42713 35.08281 35.694302 36.122917 35.6440093
#> am           0.000000  0.00000  0.00000  0.000000  0.000000  0.0000000
#> carb         0.000000  0.00000  0.00000  0.000000  0.000000  0.0000000
#> cyl          0.000000  0.00000  0.00000  0.000000  0.000000  0.0000000
#> disp         0.000000  0.00000  0.00000  0.000000  0.000000  0.0000000
#> drat         0.000000  0.00000  0.00000  0.000000  0.000000  0.0000000
#> gear         0.000000  0.00000  0.00000  0.000000  0.000000  0.0000000
#> hp           0.000000  0.00000  0.00000  0.000000  0.000000  0.0000000
#> qsec         0.000000  0.00000  0.00000  0.000000  0.000000  0.0246351
#> vs           0.000000  0.00000  0.00000  0.000000  0.000000  0.0000000
#> wt          -4.206934 -4.42335 -4.62518 -4.813408 -4.945344 -4.9354927
#>                 1.3641     1.2721     1.1864    1.1064    1.0319     0.9623
#> (Intercept) 33.8885642 32.2515508 30.7248666 29.301076 27.973244 26.7353357
#> am           0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.0000000
#> qsec         0.1149356  0.1991439  0.2776769  0.350917  0.419221  0.4828992
#> vs           0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.0000000
#> wt          -4.8993830 -4.8657095 -4.8343054 -4.805018 -4.777704 -4.7522403
#>                0.8975     0.8370     0.7806    0.7280     0.6789     0.6331
#> (Intercept) 25.580426 24.5033534 23.4988718 22.562089 21.6884419 20.8736759
#> am           0.000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> carb         0.000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> cyl          0.000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> disp         0.000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> drat         0.000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> gear         0.000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> hp           0.000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> qsec         0.542308  0.5977129  0.6493836  0.697572  0.7425126  0.7844243
#> vs           0.000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> wt          -4.728484 -4.7063282 -4.6856659 -4.666396 -4.6484252 -4.6316654
#>                 0.5905    0.5507    0.5136     0.4789     0.4467    0.4166
#> (Intercept) 20.1138224 19.405180 18.887745 18.3348461 15.7176841 13.273426
#> am           0.0000000  0.000000  0.000000  0.1725771  0.9880854  1.749678
#> carb         0.0000000  0.000000  0.000000  0.0000000  0.0000000  0.000000
#> cyl          0.0000000  0.000000  0.000000  0.0000000  0.0000000  0.000000
#> disp         0.0000000  0.000000  0.000000  0.0000000  0.0000000  0.000000
#> drat         0.0000000  0.000000  0.000000  0.0000000  0.0000000  0.000000
#> gear         0.0000000  0.000000  0.000000  0.0000000  0.0000000  0.000000
#> hp           0.0000000  0.000000  0.000000  0.0000000  0.0000000  0.000000
#> qsec         0.8235113  0.859964  0.886581  0.9024470  0.9776502  1.047888
#> vs           0.0000000  0.000000  0.000000  0.0000000  0.0000000  0.000000
#> wt          -4.6160351 -4.601458 -4.590815 -4.5269269 -4.2249359 -3.942908
#>                0.3885    0.3623    0.3379    0.3151    0.2939    0.2741
#> (Intercept) 10.997248 10.962769 10.960248 10.958686 10.957718 10.957119
#> am           2.458937  2.469667  2.470333  2.470746  2.471001  2.471159
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         1.113293  1.114285  1.114366  1.114416  1.114447  1.114467
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -3.680262 -3.676288 -3.676033 -3.675876 -3.675778 -3.675718
#>                0.2556    0.2384    0.2223    0.2073     0.1934      0.1803
#> (Intercept) 10.956748 10.956518 10.956376 10.956288 11.0199608 18.55224407
#> am           2.471257  2.471318  2.471355  2.471379  2.4817936  2.42564178
#> carb         0.000000  0.000000  0.000000  0.000000  0.0000000  0.00000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.0000000  0.00000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.0000000  0.00000000
#> drat         0.000000  0.000000  0.000000  0.000000 -0.0220201  0.00000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.0000000  0.00000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.0000000 -0.01930833
#> qsec         1.114479  1.114486  1.114490  1.114493  1.1156084  0.72881653
#> vs           0.000000  0.000000  0.000000  0.000000  0.0000000  0.00000000
#> wt          -3.675680 -3.675657 -3.675643 -3.675634 -3.6781722 -3.01063009
#>                  0.1682      0.1568      0.1463      0.1364      0.1272
#> (Intercept) 18.55491000 18.55708082 18.55884849 14.28176180 14.28227973
#> am           2.42565128  2.42565903  2.42566534  1.98093152  1.98091774
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.01750271  0.01749846
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.95565408  0.95559328
#> hp          -0.01931571 -0.01932171 -0.01932661 -0.03579525 -0.03579167
#> qsec         0.72867936  0.72856767  0.72847672  0.84321562  0.84317974
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -3.01036290 -3.01014532 -3.00996815 -3.84895142 -3.84870310
#>                  0.1186      0.1106      0.1032      0.0962      0.0897
#> (Intercept) 14.28284393 14.28343716 14.28404520 14.28465640 14.28526137
#> am           1.98089682  1.98087047  1.98084014  1.98080704  1.98077220
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.01749457  0.01749102  0.01748779  0.01748487  0.01748223
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.95554472  0.95550669  0.95547766  0.95545627  0.95544132
#> hp          -0.03578867 -0.03578620 -0.03578417 -0.03578254 -0.03578125
#> qsec         0.84314105  0.84310065  0.84305947  0.84301824  0.84297758
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -3.84846491 -3.84823826 -3.84802414 -3.84782314 -3.84763556
#>                  0.0837      0.0781      0.0728      0.0679      0.0633
#> (Intercept) 14.28585266 14.28642449 14.28697249 14.28749350 14.28798537
#> am           1.98073647  1.98070052  1.98066492  1.98063011  1.98059643
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.01747987  0.01747775  0.01747587  0.01747419  0.01747272
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.95543176  0.95542665  0.95542520  0.95542670  0.95543056
#> hp          -0.03578025 -0.03577949 -0.03577894 -0.03577857 -0.03577835
#> qsec         0.84293794  0.84289971  0.84286314  0.84282844  0.84279573
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -3.84746144 -3.84730063 -3.84715282 -3.84701755 -3.84689430
#>                  0.0590      0.0551      0.0514      0.0479      0.0447
#> (Intercept) 14.28844678 14.28887712 14.28927636 15.37717218 15.37911797
#> am           1.98056415  1.98053345  1.98050446  1.78687681  1.78661963
#> carb         0.00000000  0.00000000  0.00000000 -0.29452718 -0.29474945
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.01747142  0.01747028  0.01746928  0.01428899  0.01428556
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.95543624  0.95544331  0.95545140  1.20280102  1.20306100
#> hp          -0.03577824 -0.03577824 -0.03577831 -0.03372808 -0.03372894
#> qsec         0.84276510  0.84273656  0.84271013  0.71896812  0.71880029
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -3.84678246 -3.84668139 -3.84659041 -3.38170624 -3.38117914
#>                  0.0417      0.0388      0.0362      0.0338      0.0315
#> (Intercept) 15.38095721 15.38269552 15.38433829 15.38589060 15.38735729
#> am           1.78637654  1.78614679  1.78592969  1.78572454  1.78553072
#> carb        -0.29495937 -0.29515762 -0.29534483 -0.29552159 -0.29568849
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.01428232  0.01427927  0.01427639  0.01427367  0.01427110
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         1.20330675  1.20353901  1.20375851  1.20396593  1.20416190
#> hp          -0.03372977 -0.03373055 -0.03373130 -0.03373201 -0.03373269
#> qsec         0.71864167  0.71849179  0.71835018  0.71821639  0.71809000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -3.38068126 -3.38021104 -3.37976697 -3.37934764 -3.37895168
#>                  0.0294      0.0274      0.0256      0.0238      0.0222
#> (Intercept) 15.38874295 15.39005194 15.39128840 15.39245623 15.39355917
#> am           1.78534761  1.78517464  1.78501126  1.78485696  1.78471123
#> carb        -0.29584605 -0.29599480 -0.29613521 -0.29626774 -0.29639283
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.01426868  0.01426640  0.01426425  0.01426221  0.01426030
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         1.20434705  1.20452195  1.20468716  1.20484321  1.20499058
#> hp          -0.03373333 -0.03373395 -0.03373453 -0.03373509 -0.03373562
#> qsec         0.71797062  0.71785786  0.71775136  0.71765079  0.71755583
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -3.37857784 -3.37822489 -3.37789170 -3.37757717 -3.37728029
#>                  0.0207      0.0193      0.0180      0.0168      0.0157
#> (Intercept) 15.39460072 15.39558423 15.39651287 15.39738963 15.39821734
#> am           1.78457361  1.78444367  1.78432099  1.78420516  1.78409581
#> carb        -0.29651088 -0.29662230 -0.29672743 -0.29682664 -0.29692025
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.01425849  0.01425679  0.01425518  0.01425366  0.01425224
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         1.20512975  1.20526117  1.20538525  1.20550241  1.20561301
#> hp          -0.03373612 -0.03373660 -0.03373705 -0.03373749 -0.03373789
#> qsec         0.71746616  0.71738151  0.71730158  0.71722614  0.71715492
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -3.37700008 -3.37673562 -3.37648605 -3.37625053 -3.37602829
#>                  0.0146      0.0136      0.0127      0.0119      0.0111
#> (Intercept) 15.39899871 15.39973627 15.40043244 15.40108950 15.40170960
#> am           1.78399258  1.78389515  1.78380319  1.78371639  1.78363448
#> carb        -0.29700858 -0.29709191 -0.29717052 -0.29724469 -0.29731465
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.01425089  0.01424962  0.01424842  0.01424729  0.01424622
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         1.20571741  1.20581597  1.20590899  1.20599679  1.20607965
#> hp          -0.03373828 -0.03373865 -0.03373900 -0.03373933 -0.03373965
#> qsec         0.71708770  0.71702426  0.71696438  0.71690787  0.71685455
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -3.37581859 -3.37562073 -3.37543406 -3.37525795 -3.37509181
#>                  0.0103      0.0096      0.0090      0.0084      0.0078
#> (Intercept) 16.85519135 16.85838251 16.86139574 16.86423860 16.86691842
#> am           1.81884407  1.81899395  1.81913055  1.81925461  1.81936684
#> carb        -0.20068400 -0.20049966 -0.20033342 -0.20018422 -0.20005107
#> cyl         -0.12021704 -0.12048292 -0.12073086 -0.12096172 -0.12117634
#> disp         0.01492855  0.01492643  0.01492428  0.01492211  0.01491992
#> drat        -0.25644534 -0.25705312 -0.25762163 -0.25815273 -0.25864823
#> gear         1.15109615  1.15087248  1.15066704  1.15047882  1.15030690
#> hp          -0.03236866 -0.03236015 -0.03235209 -0.03234444 -0.03233719
#> qsec         0.72670237  0.72673346  0.72675935  0.72678038  0.72679688
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -3.49453978 -3.49466122 -3.49475957 -3.49483640 -3.49489321
#>                  0.0073      0.0068      0.0063      0.0059       0.0055
#> (Intercept) 16.86944230 16.87181709 16.87404944 16.87614576 16.878066751
#> am           1.81946793  1.81955851  1.81963922  1.81971064  1.819575399
#> carb        -0.19993300 -0.19982910 -0.19973850 -0.19966035 -0.199343405
#> cyl         -0.12137554 -0.12156010 -0.12173076 -0.12188825 -0.122059518
#> disp         0.01491773  0.01491553  0.01491333  0.01491113  0.014906729
#> drat        -0.25910984 -0.25953923 -0.25993801 -0.26030771 -0.261033540
#> gear         1.15015038  1.15000838  1.14988009  1.14976472  1.149586445
#> hp          -0.03233033 -0.03232384 -0.03231770 -0.03231191 -0.032301398
#> qsec         0.72680916  0.72681752  0.72682223  0.72682358  0.726772180
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.001997388
#> wt          -3.49493139 -3.49495231 -3.49495723 -3.49494737 -3.494641607
#>                  0.0051      0.0048
#> (Intercept) 16.79457181 16.79167009
#> am           1.85439177  1.85518285
#> carb        -0.19268865 -0.19241955
#> cyl         -0.10351831 -0.10328516
#> disp         0.01490130  0.01490461
#> drat        -0.27570794 -0.27604164
#> gear         1.13530325  1.13492268
#> hp          -0.03226827 -0.03226639
#> qsec         0.72694426  0.72712892
#> vs           0.08979421  0.09081287
#> wt          -3.49661413 -3.49716732
#> 
#> $iter
#>   [1]   0   3   3   3   2   2   2   2   2   2   2   2   2   2   2   2   2   4
#>  [19]   4   4   4   4   4   3   3   3   3   3   3   3   3   3   3  26  37  37
#>  [37]  36   4   1   1   1   1   1   1   1   1   8  36   1   1   1 118   1   1
#>  [55]   1   1   1   1   1   1   1   1   1   1   1   1  92   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1  64   1
#>  [91]   1   1   1   1   1   1   1   2  25   1
#> 
#> $lambda
#>   [1] 4.789467731 4.466673636 4.165634782 3.884884940 3.623056697 3.378874802
#>   [7] 3.151149949 2.938772989 2.740709525 2.555994876 2.383729376 2.223073994
#>  [13] 2.073246247 1.933516389 1.803203856 1.681673951 1.568334755 1.462634242
#>  [19] 1.364057589 1.272124673 1.186387727 1.106429165 1.031859543 0.962315663
#>  [25] 0.897458808 0.836973088 0.780563903 0.727956509 0.678894676 0.633139447
#>  [31] 0.590467968 0.550672403 0.513558926 0.478946773 0.446667364 0.416563478
#>  [37] 0.388488494 0.362305670 0.337887480 0.315114995 0.293877299 0.274070953
#>  [43] 0.255599488 0.238372938 0.222307399 0.207324625 0.193351639 0.180320386
#>  [49] 0.168167395 0.156833475 0.146263424 0.136405759 0.127212467 0.118638773
#>  [55] 0.110642917 0.103185954 0.096231566 0.089745881 0.083697309 0.078056390
#>  [61] 0.072795651 0.067889468 0.063313945 0.059046797 0.055067240 0.051355893
#>  [67] 0.047894677 0.044666736 0.041656348 0.038848849 0.036230567 0.033788748
#>  [73] 0.031511499 0.029387730 0.027407095 0.025559949 0.023837294 0.022230740
#>  [79] 0.020732462 0.019335164 0.018032039 0.016816740 0.015683348 0.014626342
#>  [85] 0.013640576 0.012721247 0.011863877 0.011064292 0.010318595 0.009623157
#>  [91] 0.008974588 0.008369731 0.007805639 0.007279565 0.006788947 0.006331394
#>  [97] 0.005904680 0.005506724 0.005135589 0.004789468
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 33
#> 
#> $loss
#>   [1] 650.29143 557.81594 480.44645 416.00907 362.62700 318.68085 282.77423
#>   [8] 253.70394 230.43400 212.07317 197.85545 187.12315 179.31220 173.93940
#>  [15] 170.59144 168.91527 168.57240 165.88990 156.70348 149.05195 142.71201
#>  [22] 137.49160 133.22511 129.77093 127.00372 124.81916 123.12638 121.84736
#>  [29] 120.91517 120.27250 119.87030 119.66669 119.62256 118.01958 112.01376
#>  [36] 108.74376 107.72978 107.72948 107.72948 107.72947 107.72947 107.72947
#>  [43] 107.72947 107.72947 107.72947 107.72947 107.64889 103.09227 103.09227
#>  [50] 103.09226 103.09226  94.97170  94.97169  94.97169  94.97168  94.97168
#>  [57]  94.97167  94.97167  94.97167  94.97167  94.97167  94.97166  94.97166
#>  [64]  94.97166  94.97166  94.97166  94.60474  94.60472  94.60471  94.60470
#>  [71]  94.60469  94.60468  94.60468  94.60467  94.60466  94.60466  94.60465
#>  [78]  94.60465  94.60465  94.60464  94.60464  94.60464  94.60463  94.60463
#>  [85]  94.60463  94.60463  94.60463  94.60463  94.53269  94.53268  94.53266
#>  [92]  94.53265  94.53264  94.53263  94.53262  94.53262  94.53261  94.53236
#>  [99]  94.52695  94.52692
#> 
#> $linear.predictors
#>      4.7895   4.4667   4.1656   3.8849   3.6231   3.3789   3.1511   2.9388
#> 1  20.05714 20.37144 20.66456 20.93793 21.19287 21.43062 21.65236 21.85915
#> 2  20.05714 20.52143 20.95442 21.35824 21.73483 22.08605 22.41359 22.71906
#> 3  20.05714 20.07397 20.08967 20.10431 20.11796 20.13070 20.14257 20.15364
#> 4  20.05714 19.96149 19.87228 19.78908 19.71149 19.63913 19.57164 19.50871
#> 5  20.05714 20.08647 20.11383 20.13934 20.16313 20.18531 20.20601 20.22530
#> 6  20.05714 20.10647 20.15247 20.19538 20.23539 20.27271 20.30751 20.33996
#> 7  20.05714 19.96149 19.87228 19.78908 19.71149 19.63913 19.57164 19.50871
#> 8  20.05714 19.96149 19.87228 19.78908 19.71149 19.63913 19.57164 19.50871
#> 9  20.05714 19.64652 19.26357 18.90643 18.57336 18.26273 17.97305 17.70288
#> 10 20.05714 19.81650 19.59208 19.38278 19.18759 19.00555 18.83578 18.67745
#> 11 20.05714 19.79150 19.54377 19.31273 19.09726 18.89631 18.70891 18.53413
#> 12 20.05714 18.96959 17.95533 17.00943 16.12728 15.30458 14.53733 13.82179
#> 13 20.05714 19.00908 18.03166 17.12011 16.26999 15.47717 14.73779 14.04823
#> 14 20.05714 20.76390 21.42303 22.03774 22.61101 23.14565 23.64426 24.10926
#> 15 20.05714 20.44894 20.81432 21.15509 21.47288 21.76926 22.04566 22.30344
#> 16 20.05714 19.92149 19.79498 19.67700 19.56696 19.46435 19.36865 19.27940
#> 17 20.05714 19.76151 19.48579 19.22866 18.98886 18.76523 18.55666 18.36215
#> 18 20.05714 20.71391 21.32641 21.89764 22.43036 22.92718 23.39052 23.82262
#> 19 20.05714 20.92489 21.73415 22.48887 23.19273 23.84914 24.46132 25.03224
#> 20 20.05714 20.09647 20.13315 20.16736 20.19926 20.22901 20.25676 20.28263
#> 21 20.05714 20.29145 20.50997 20.71376 20.90382 21.08106 21.24636 21.40052
#>      2.7407   2.5560   2.3837   2.2231   2.0732   1.9335    1.8032    1.6817
#> 1  22.05200 22.23185 22.39959 22.55602 22.70190 22.83796 22.964840 23.083172
#> 2  23.00395 23.26963 23.51740 23.74848 23.96398 24.16496 24.352394 24.527195
#> 3  20.16397 20.17360 20.18259 20.19096 20.19878 20.20606 20.212857 20.219194
#> 4  19.45001 19.39527 19.34422 19.29662 19.25222 19.21081 19.172192 19.136177
#> 5  20.24330 20.26009 20.27574 20.29034 20.30395 20.31665 20.328487 20.339529
#> 6  20.37023 20.39845 20.42478 20.44933 20.47223 20.49358 20.513494 20.532066
#> 7  19.45001 19.39527 19.34422 19.29662 19.25222 19.21081 19.172192 19.136177
#> 8  19.45001 19.39527 19.34422 19.29662 19.25222 19.21081 19.172192 19.136177
#> 9  17.45093 17.21595 16.99681 16.79244 16.60185 16.42410 16.258328 16.103730
#> 10 18.52980 18.39209 18.26367 18.14390 18.03221 17.92804 17.830889 17.740289
#> 11 18.37114 18.21913 18.07737 17.94516 17.82186 17.70687 17.599630 17.499619
#> 12 13.15447 12.53213 11.95174 11.41046 10.90566 10.43488  9.995834  9.586375
#> 13 13.40515 12.80541 12.24610 11.72447 11.23801 10.78433 10.361223  9.966635
#> 14 24.54293 24.94736 25.32454 25.67630 26.00435 26.31029 26.595606 26.861698
#> 15 22.54384 22.76804 22.97713 23.17212 23.35398 23.52357 23.681743 23.829250
#> 16 19.19616 19.11853 19.04614 18.97863 18.91566 18.85694 18.802177 18.751105
#> 17 18.18075 18.01158 17.85380 17.70667 17.56944 17.44147 17.322119 17.210814
#> 18 24.22561 24.60144 24.95193 25.27881 25.58365 25.86795 26.133088 26.380357
#> 19 25.56468 26.06124 26.52433 26.95621 27.35898 27.73460 28.084914 28.411615
#> 20 20.30676 20.32927 20.35026 20.36983 20.38809 20.40511 20.420990 20.435798
#> 21 21.54429 21.67838 21.80342 21.92004 22.02879 22.13022 22.224811 22.313027
#>       1.5683    1.4626    1.3641    1.2721    1.1864    1.1064    1.0319
#> 1  23.166116 23.118512 22.944020 22.781301 22.629548 22.488023 22.356036
#> 2  24.649719 24.652125 24.660947 24.669173 24.676845 24.684000 24.690672
#> 3  20.223636 20.255307 20.371396 20.479653 20.580613 20.674770 20.762580
#> 4  19.110934 19.085204 18.990890 18.902940 18.820917 18.744422 18.673082
#> 5  20.347270 20.392490 20.558244 20.712816 20.856970 20.991409 21.116787
#> 6  20.545083 20.661351 21.087533 21.484962 21.855605 22.201269 22.523636
#> 7  19.110934 19.116737 19.138008 19.157844 19.176343 19.193595 19.209685
#> 8  19.110934 19.131518 19.206969 19.277330 19.342949 19.404145 19.461217
#> 9  15.995367 15.985205 15.947954 15.913217 15.880821 15.850609 15.822432
#> 10 17.676784 17.668199 17.636732 17.607387 17.580021 17.554498 17.530696
#> 11 17.429517 17.431279 17.437737 17.443759 17.449376 17.454614 17.459499
#> 12  9.299371  9.312894  9.362463  9.408687  9.451796  9.492000  9.529494
#> 13  9.690054  9.692944  9.703540  9.713421  9.722635  9.731229  9.739244
#> 14 27.048211 27.077619 27.185414 27.285938 27.379686 27.467116 27.548654
#> 15 23.932644 23.970968 24.111446 24.242447 24.364618 24.478556 24.584814
#> 16 18.715306 18.686669 18.581699 18.483811 18.392521 18.307383 18.227983
#> 17 17.132796 17.071344 16.846091 16.636034 16.440135 16.257438 16.087055
#> 18 26.553676 26.559434 26.580541 26.600223 26.618579 26.635697 26.651662
#> 19 28.640611 28.592942 28.418209 28.255264 28.103302 27.961581 27.829412
#> 20 20.446177 20.355706 20.024086 19.714838 19.426433 19.157466 18.906625
#> 21 22.374861 22.381552 22.406081 22.428955 22.450288 22.470182 22.488736
#>       0.9623    0.8975    0.8370    0.7806    0.7280    0.6789    0.6331
#> 1  22.232987 22.118189 22.011128 21.911282 21.818165 21.731325 21.650337
#> 2  24.696893 24.702696 24.708109 24.713156 24.717864 24.722254 24.726348
#> 3  20.844444 20.920819 20.992047 21.058474 21.120424 21.178199 21.232080
#> 4  18.606574 18.544525 18.486658 18.432691 18.382361 18.335423 18.291649
#> 5  21.233674 21.342724 21.444424 21.539270 21.627724 21.710217 21.787149
#> 6  22.824171 23.104556 23.366045 23.609910 23.837339 24.049440 24.247246
#> 7  19.224685 19.238679 19.251730 19.263902 19.275253 19.285839 19.295712
#> 8  19.514424 19.564064 19.610358 19.653532 19.693796 19.731347 19.766366
#> 9  15.796164 15.771657 15.748802 15.727487 15.707609 15.689070 15.671781
#> 10 17.508506 17.487803 17.468496 17.450490 17.433698 17.418037 17.403432
#> 11 17.464053 17.468302 17.472265 17.475960 17.479407 17.482621 17.485618
#> 12  9.564448  9.597060  9.627473  9.655836  9.682288  9.706957  9.729964
#> 13  9.746716  9.753687  9.760188  9.766251  9.771905  9.777178  9.782096
#> 14 27.624669 27.695588 27.761728 27.823409 27.880934 27.934582 27.984613
#> 15 24.683877 24.776297 24.862489 24.942872 25.017837 25.087750 25.152951
#> 16 18.153960 18.084900 18.020495 17.960430 17.904413 17.852172 17.803452
#> 17 15.928210 15.780015 15.641809 15.512917 15.392711 15.280608 15.176059
#> 18 26.666546 26.680432 26.693382 26.705459 26.716722 26.727226 26.737023
#> 19 27.706193 27.591236 27.484027 27.384043 27.290798 27.203837 27.122737
#> 20 18.672773 18.454599 18.251130 18.061374 17.884406 17.719366 17.565449
#> 21 22.506033 22.522171 22.537221 22.551256 22.564346 22.576553 22.587938
#>       0.5905    0.5507    0.5136    0.4789   0.4467   0.4166   0.3885   0.3623
#> 1  21.574807 21.504368 21.452934 21.501152 21.72856 21.94092 22.13870 22.14169
#> 2  24.730167 24.733728 24.736328 24.799491 25.09799 25.37675 25.63636 25.64029
#> 3  21.282330 21.329193 21.363411 21.324345 21.14003 20.96792 20.80762 20.80520
#> 4  18.250824 18.212752 18.184952 18.121865 17.82351 17.54487 17.28539 17.28146
#> 5  21.858897 21.925809 21.974667 21.942889 21.79314 21.65331 21.52307 21.52110
#> 6  24.431721 24.603763 24.729385 24.741062 24.79733 24.84990 24.89883 24.89958
#> 7  19.304919 19.313506 19.319776 19.276997 19.07490 18.88617 18.71041 18.70775
#> 8  19.799026 19.829484 19.851724 19.818465 19.66149 19.51490 19.37838 19.37632
#> 9  15.655657 15.640619 15.629640 15.612831 15.53331 15.45904 15.38988 15.38883
#> 10 17.389811 17.377108 17.367833 17.332476 17.16532 17.00921 16.86383 16.86163
#> 11 17.488414 17.491021 17.492924 17.467108 17.34513 17.23122 17.12513 17.12353
#> 12  9.751420  9.771430  9.786041  9.862400 10.22336 10.56046 10.87439 10.87914
#> 13  9.786682  9.790959  9.794083  9.859048 10.16607 10.45279 10.71981 10.72385
#> 14 28.031273 28.074789 28.106563 28.159207 28.40825 28.64084 28.85743 28.86071
#> 15 25.213758 25.270466 25.311873 25.233935 24.86600 24.52239 24.20239 24.19756
#> 16 17.758015 17.715641 17.684700 17.624344 17.33887 17.07226 16.82398 16.82022
#> 17 15.078557 14.987626 14.921231 14.858155 14.55952 14.28061 14.02089 14.01695
#> 18 26.746159 26.754679 26.760900 26.804067 27.00811 27.19866 27.37611 27.37880
#> 19 27.047103 26.976566 26.925062 26.909537 26.83573 26.76679 26.70260 26.70162
#> 20 17.421905 17.288036 17.190288 17.242546 17.48865 17.71846 17.93250 17.93573
#> 21 22.598556 22.608457 22.615688 22.708080 23.14474 23.55253 23.93230 23.93805
#>      0.3379   0.3151   0.2939   0.2741   0.2556   0.2384   0.2223   0.2073
#> 1  22.14184 22.14193 22.14198 22.14202 22.14204 22.14205 22.14206 22.14207
#> 2  25.64053 25.64068 25.64078 25.64084 25.64087 25.64090 25.64091 25.64092
#> 3  20.80507 20.80500 20.80495 20.80492 20.80490 20.80489 20.80488 20.80488
#> 4  17.28120 17.28104 17.28093 17.28087 17.28083 17.28081 17.28079 17.28078
#> 5  21.52102 21.52097 21.52093 21.52091 21.52090 21.52089 21.52089 21.52088
#> 6  24.89972 24.89981 24.89986 24.89989 24.89991 24.89993 24.89993 24.89994
#> 7  18.70759 18.70749 18.70743 18.70739 18.70737 18.70735 18.70734 18.70734
#> 8  19.37621 19.37614 19.37610 19.37607 19.37605 19.37604 19.37604 19.37603
#> 9  15.38876 15.38871 15.38868 15.38867 15.38866 15.38865 15.38865 15.38864
#> 10 16.86148 16.86139 16.86134 16.86130 16.86128 16.86127 16.86126 16.86126
#> 11 17.12343 17.12337 17.12333 17.12331 17.12329 17.12328 17.12328 17.12327
#> 12 10.87944 10.87963 10.87975 10.87982 10.87987 10.87989 10.87991 10.87992
#> 13 10.72410 10.72426 10.72436 10.72442 10.72445 10.72448 10.72449 10.72450
#> 14 28.86094 28.86108 28.86117 28.86122 28.86125 28.86128 28.86129 28.86130
#> 15 24.19729 24.19712 24.19702 24.19695 24.19691 24.19689 24.19687 24.19686
#> 16 16.81996 16.81980 16.81971 16.81964 16.81961 16.81958 16.81957 16.81956
#> 17 14.01666 14.01648 14.01636 14.01629 14.01625 14.01622 14.01621 14.01620
#> 18 27.37897 27.37908 27.37914 27.37918 27.37921 27.37922 27.37923 27.37924
#> 19 26.70153 26.70146 26.70143 26.70140 26.70139 26.70138 26.70137 26.70137
#> 20 17.93586 17.93594 17.93599 17.93602 17.93604 17.93605 17.93606 17.93606
#> 21 23.93841 23.93864 23.93878 23.93886 23.93892 23.93895 23.93897 23.93898
#>      0.1934   0.1803   0.1682   0.1568   0.1463   0.1364   0.1272   0.1186
#> 1  22.14198 22.96244 22.96275 22.96299 22.96320 22.74334 22.74338 22.74342
#> 2  25.64509 25.76083 25.76088 25.76093 25.76097 25.40932 25.40936 25.40940
#> 3  20.81424 20.91735 20.91739 20.91743 20.91746 21.74468 21.74441 21.74417
#> 4  17.28534 17.22118 17.22114 17.22110 17.22108 18.29667 18.29634 18.29605
#> 5  21.51750 22.32755 22.32787 22.32812 22.32833 23.03888 23.03883 23.03880
#> 6  24.89483 23.92437 23.92403 23.92376 23.92354 24.35365 24.35363 24.35360
#> 7  18.69636 19.15809 19.15826 19.15840 19.15851 18.82547 18.82567 18.82587
#> 8  19.36573 19.59538 19.59547 19.59554 19.59560 19.33140 19.33158 19.33175
#> 9  15.39378 15.50489 15.50493 15.50496 15.50498 14.53955 14.53974 14.53992
#> 10 16.86748 16.67427 16.67419 16.67412 16.67407 16.01683 16.01694 16.01703
#> 11 17.12982 16.81526 16.81514 16.81504 16.81496 16.16167 16.16177 16.16186
#> 12 10.88364 11.05881 11.05889 11.05896 11.05902 11.65338 11.65324 11.65311
#> 13 10.72290 10.71549 10.71550 10.71551 10.71551 10.73318 10.73317 10.73317
#> 14 28.85999 28.70179 28.70174 28.70171 28.70168 28.72023 28.72016 28.72009
#> 15 24.19512 23.84175 23.84162 23.84151 23.84142 23.16374 23.16381 23.16386
#> 16 16.83233 17.35371 17.35390 17.35405 17.35417 18.02204 18.02183 18.02164
#> 17 14.00517 13.49195 13.49172 13.49153 13.49138 12.71881 12.71894 12.71905
#> 18 27.37965 27.65260 27.65271 27.65280 27.65288 27.59459 27.59456 27.59452
#> 19 26.70745 26.55796 26.55789 26.55783 26.55778 27.08749 27.08746 27.08743
#> 20 17.92535 16.90463 16.90422 16.90388 16.90360 17.75992 17.75984 17.75976
#> 21 23.93625 24.05971 24.05978 24.05983 24.05987 23.28518 23.28534 23.28549
#>      0.1106   0.1032   0.0962   0.0897   0.0837   0.0781   0.0728   0.0679
#> 1  22.74347 22.74352 22.74357 22.74362 22.74368 22.74373 22.74378 22.74383
#> 2  25.40944 25.40947 25.40950 25.40952 25.40954 25.40956 25.40958 25.40959
#> 3  21.74395 21.74375 21.74356 21.74340 21.74325 21.74312 21.74300 21.74290
#> 4  18.29577 18.29552 18.29529 18.29508 18.29489 18.29472 18.29457 18.29443
#> 5  23.03878 23.03879 23.03880 23.03883 23.03886 23.03889 23.03893 23.03898
#> 6  24.35356 24.35353 24.35348 24.35344 24.35340 24.35335 24.35331 24.35327
#> 7  18.82606 18.82624 18.82641 18.82658 18.82673 18.82688 18.82701 18.82713
#> 8  19.33192 19.33208 19.33223 19.33236 19.33249 19.33262 19.33273 19.33283
#> 9  14.54009 14.54024 14.54037 14.54049 14.54060 14.54070 14.54079 14.54087
#> 10 16.01711 16.01718 16.01723 16.01728 16.01733 16.01736 16.01739 16.01742
#> 11 16.16194 16.16200 16.16205 16.16209 16.16213 16.16216 16.16218 16.16220
#> 12 11.65300 11.65290 11.65281 11.65273 11.65266 11.65259 11.65254 11.65249
#> 13 10.73316 10.73315 10.73314 10.73314 10.73313 10.73312 10.73312 10.73311
#> 14 28.72003 28.71997 28.71990 28.71985 28.71979 28.71974 28.71969 28.71965
#> 15 23.16390 23.16394 23.16396 23.16398 23.16399 23.16400 23.16400 23.16400
#> 16 18.02148 18.02134 18.02121 18.02110 18.02101 18.02093 18.02086 18.02080
#> 17 12.71914 12.71921 12.71928 12.71932 12.71936 12.71939 12.71941 12.71943
#> 18 27.59450 27.59447 27.59445 27.59443 27.59442 27.59441 27.59439 27.59438
#> 19 27.08741 27.08740 27.08738 27.08737 27.08736 27.08735 27.08734 27.08734
#> 20 17.75968 17.75959 17.75951 17.75942 17.75934 17.75927 17.75919 17.75912
#> 21 23.28562 23.28574 23.28585 23.28595 23.28604 23.28611 23.28618 23.28624
#>      0.0633   0.0590   0.0551   0.0514   0.0479   0.0447   0.0417   0.0388
#> 1  22.74388 22.74392 22.74396 22.74400 22.34744 22.34725 22.34708 22.34691
#> 2  25.40960 25.40962 25.40962 25.40963 25.62166 25.62182 25.62197 25.62210
#> 3  21.74280 21.74272 21.74265 21.74259 21.77207 21.77203 21.77199 21.77195
#> 4  18.29431 18.29420 18.29411 18.29402 18.24191 18.24176 18.24162 18.24149
#> 5  23.03902 23.03906 23.03911 23.03915 23.19610 23.19640 23.19670 23.19697
#> 6  24.35323 24.35319 24.35315 24.35312 24.21904 24.21883 24.21864 24.21845
#> 7  18.82724 18.82735 18.82744 18.82753 18.78060 18.78075 18.78090 18.78104
#> 8  19.33292 19.33301 19.33308 19.33315 19.21198 19.21203 19.21209 19.21214
#> 9  14.54094 14.54100 14.54105 14.54110 14.71834 14.71853 14.71870 14.71886
#> 10 16.01744 16.01746 16.01747 16.01748 16.01192 16.01189 16.01186 16.01183
#> 11 16.16221 16.16222 16.16223 16.16223 16.13042 16.13035 16.13028 16.13022
#> 12 11.65245 11.65241 11.65238 11.65235 11.59850 11.59844 11.59839 11.59834
#> 13 10.73310 10.73310 10.73309 10.73309 10.78637 10.78639 10.78641 10.78643
#> 14 28.71961 28.71957 28.71953 28.71950 28.60638 28.60622 28.60606 28.60591
#> 15 23.16400 23.16399 23.16399 23.16398 23.18618 23.18613 23.18608 23.18603
#> 16 18.02074 18.02070 18.02067 18.02064 18.10659 18.10668 18.10676 18.10684
#> 17 12.71944 12.71945 12.71945 12.71945 12.63878 12.63864 12.63851 12.63839
#> 18 27.59438 27.59437 27.59436 27.59436 27.62840 27.62843 27.62845 27.62847
#> 19 27.08733 27.08733 27.08733 27.08733 27.17065 27.17073 27.17081 27.17088
#> 20 17.75906 17.75900 17.75894 17.75889 17.81620 17.81610 17.81601 17.81593
#> 21 23.28630 23.28634 23.28639 23.28642 23.41047 23.41059 23.41070 23.41081
#>      0.0362   0.0338   0.0315   0.0294   0.0274   0.0256   0.0238   0.0222
#> 1  22.34676 22.34661 22.34647 22.34634 22.34622 22.34610 22.34599 22.34589
#> 2  25.62223 25.62236 25.62247 25.62258 25.62269 25.62278 25.62288 25.62296
#> 3  21.77191 21.77188 21.77184 21.77181 21.77178 21.77176 21.77173 21.77171
#> 4  18.24137 18.24125 18.24114 18.24104 18.24094 18.24085 18.24076 18.24068
#> 5  23.19724 23.19748 23.19772 23.19794 23.19815 23.19834 23.19853 23.19871
#> 6  24.21828 24.21811 24.21796 24.21781 24.21767 24.21754 24.21741 24.21730
#> 7  18.78118 18.78130 18.78142 18.78153 18.78164 18.78174 18.78183 18.78192
#> 8  19.21219 19.21223 19.21227 19.21231 19.21235 19.21239 19.21242 19.21246
#> 9  14.71901 14.71916 14.71930 14.71943 14.71955 14.71966 14.71977 14.71987
#> 10 16.01181 16.01178 16.01176 16.01174 16.01171 16.01170 16.01168 16.01166
#> 11 16.13016 16.13010 16.13005 16.13000 16.12995 16.12990 16.12986 16.12982
#> 12 11.59829 11.59824 11.59820 11.59816 11.59812 11.59808 11.59805 11.59802
#> 13 10.78645 10.78647 10.78649 10.78650 10.78652 10.78653 10.78655 10.78656
#> 14 28.60577 28.60564 28.60551 28.60539 28.60528 28.60518 28.60508 28.60498
#> 15 23.18599 23.18595 23.18591 23.18587 23.18584 23.18580 23.18577 23.18574
#> 16 18.10691 18.10698 18.10704 18.10710 18.10716 18.10722 18.10727 18.10732
#> 17 12.63827 12.63816 12.63806 12.63796 12.63787 12.63778 12.63770 12.63762
#> 18 27.62850 27.62852 27.62854 27.62856 27.62857 27.62859 27.62861 27.62862
#> 19 27.17095 27.17101 27.17108 27.17113 27.17119 27.17124 27.17129 27.17134
#> 20 17.81585 17.81577 17.81570 17.81563 17.81557 17.81551 17.81545 17.81539
#> 21 23.41090 23.41100 23.41109 23.41117 23.41125 23.41132 23.41139 23.41146
#>      0.0207   0.0193   0.0180   0.0168   0.0157   0.0146   0.0136   0.0127
#> 1  22.34579 22.34570 22.34561 22.34553 22.34545 22.34538 22.34531 22.34524
#> 2  25.62305 25.62312 25.62320 25.62326 25.62333 25.62339 25.62345 25.62350
#> 3  21.77168 21.77166 21.77164 21.77162 21.77161 21.77159 21.77157 21.77156
#> 4  18.24060 18.24052 18.24046 18.24039 18.24033 18.24027 18.24022 18.24016
#> 5  23.19887 23.19903 23.19918 23.19932 23.19945 23.19958 23.19969 23.19981
#> 6  24.21719 24.21708 24.21698 24.21689 24.21680 24.21672 24.21664 24.21657
#> 7  18.78201 18.78209 18.78216 18.78223 18.78230 18.78236 18.78242 18.78248
#> 8  19.21249 19.21252 19.21254 19.21257 19.21259 19.21262 19.21264 19.21266
#> 9  14.71997 14.72006 14.72014 14.72023 14.72030 14.72037 14.72044 14.72051
#> 10 16.01164 16.01163 16.01161 16.01160 16.01158 16.01157 16.01156 16.01155
#> 11 16.12978 16.12974 16.12971 16.12967 16.12964 16.12961 16.12959 16.12956
#> 12 11.59798 11.59796 11.59793 11.59790 11.59788 11.59786 11.59784 11.59782
#> 13 10.78657 10.78658 10.78659 10.78660 10.78661 10.78662 10.78663 10.78664
#> 14 28.60489 28.60481 28.60473 28.60466 28.60459 28.60452 28.60446 28.60440
#> 15 23.18571 23.18569 23.18566 23.18564 23.18561 23.18559 23.18557 23.18555
#> 16 18.10736 18.10741 18.10745 18.10749 18.10753 18.10756 18.10760 18.10763
#> 17 12.63754 12.63747 12.63740 12.63734 12.63728 12.63723 12.63717 12.63712
#> 18 27.62863 27.62865 27.62866 27.62867 27.62868 27.62869 27.62870 27.62871
#> 19 27.17138 27.17142 27.17146 27.17150 27.17153 27.17156 27.17159 27.17162
#> 20 17.81534 17.81529 17.81525 17.81521 17.81516 17.81513 17.81509 17.81505
#> 21 23.41152 23.41158 23.41164 23.41169 23.41174 23.41178 23.41183 23.41187
#>      0.0119   0.0111   0.0103   0.0096   0.0090   0.0084   0.0078   0.0073
#> 1  22.34518 22.34512 22.38809 22.38810 22.38810 22.38809 22.38808 22.38806
#> 2  25.62355 25.62360 25.62815 25.62824 25.62832 25.62841 25.62849 25.62858
#> 3  21.77154 21.77153 21.77981 21.77964 21.77949 21.77934 21.77919 21.77905
#> 4  18.24012 18.24007 18.19459 18.19428 18.19397 18.19368 18.19340 18.19313
#> 5  23.19991 23.20001 23.20068 23.20049 23.20032 23.20015 23.20000 23.19986
#> 6  24.21650 24.21643 24.23268 24.23273 24.23278 24.23283 24.23286 24.23290
#> 7  18.78253 18.78258 18.72839 18.72829 18.72819 18.72811 18.72804 18.72797
#> 8  19.21268 19.21270 19.16441 19.16433 19.16425 19.16418 19.16412 19.16406
#> 9  14.72057 14.72062 14.67019 14.67026 14.67033 14.67041 14.67050 14.67058
#> 10 16.01153 16.01152 16.00367 16.00379 16.00390 16.00401 16.00412 16.00422
#> 11 16.12953 16.12951 16.11962 16.11975 16.11987 16.11998 16.12009 16.12020
#> 12 11.59780 11.59778 11.67800 11.67805 11.67810 11.67814 11.67817 11.67819
#> 13 10.78664 10.78665 10.82030 10.82038 10.82046 10.82053 10.82060 10.82066
#> 14 28.60434 28.60429 28.52102 28.52082 28.52063 28.52045 28.52028 28.52012
#> 15 23.18553 23.18552 23.25851 23.25880 23.25907 23.25932 23.25955 23.25977
#> 16 18.10766 18.10768 18.08826 18.08804 18.08783 18.08764 18.08745 18.08728
#> 17 12.63708 12.63703 12.66159 12.66181 12.66201 12.66220 12.66238 12.66254
#> 18 27.62872 27.62873 27.56634 27.56617 27.56601 27.56587 27.56573 27.56561
#> 19 27.17165 27.17168 27.33656 27.33690 27.33721 27.33751 27.33778 27.33804
#> 20 17.81502 17.81499 17.73693 17.73677 17.73661 17.73647 17.73633 17.73620
#> 21 23.41191 23.41194 23.42221 23.42237 23.42253 23.42269 23.42284 23.42298
#>      0.0068   0.0063   0.0059   0.0055   0.0051   0.0048
#> 1  22.38803 22.38800 22.38797 22.38686 22.36212 22.36199
#> 2  25.62866 25.62874 25.62882 25.62898 25.63641 25.63640
#> 3  21.77892 21.77879 21.77866 21.77910 21.81250 21.81299
#> 4  18.19287 18.19262 18.19239 18.19118 18.18024 18.18031
#> 5  23.19973 23.19962 23.19951 23.19994 23.17884 23.17821
#> 6  24.23292 24.23294 24.23296 24.23343 24.21066 24.21055
#> 7  18.72791 18.72786 18.72782 18.72909 18.75613 18.75615
#> 8  19.16401 19.16396 19.16392 19.16515 19.19229 19.19243
#> 9  14.67067 14.67076 14.67085 14.67062 14.66695 14.66676
#> 10 16.00432 16.00441 16.00450 16.00415 16.00118 16.00123
#> 11 16.12030 16.12039 16.12048 16.12012 16.11713 16.11722
#> 12 11.67821 11.67822 11.67823 11.67831 11.67989 11.68000
#> 13 10.82072 10.82078 10.82083 10.82099 10.81988 10.81985
#> 14 28.51996 28.51981 28.51967 28.51922 28.52167 28.52187
#> 15 23.25998 23.26017 23.26035 23.26090 23.25002 23.24987
#> 16 18.08712 18.08696 18.08682 18.08585 18.07984 18.07979
#> 17 12.66270 12.66284 12.66298 12.66297 12.65813 12.65813
#> 18 27.56549 27.56538 27.56528 27.56499 27.56912 27.56915
#> 19 27.33828 27.33851 27.33871 27.33918 27.34218 27.34221
#> 20 17.73608 17.73596 17.73586 17.73504 17.73105 17.73115
#> 21 23.42312 23.42326 23.42339 23.42393 23.43375 23.43374
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 22.8 21.4 18.7 24.4 22.8 19.2 17.8 16.4 17.3 15.2 10.4 14.7 33.9 21.5
#> [16] 15.5 13.3 27.3 30.4 15.8 21.4
#> 
#> $X
#>               am       carb         cyl       disp       drat       gear
#>  [1,]  1.4142136  1.2126781 -0.05299989 -0.5597113  0.6777018  0.5834600
#>  [2,]  1.4142136 -1.3339459 -1.16599767 -1.0021645  0.5672069  0.5834600
#>  [3,] -0.7071068 -1.3339459 -0.05299989  0.2741427 -1.1344138 -0.9481224
#>  [4,] -0.7071068 -0.4850713  1.05999788  1.1420315 -0.9797210 -0.9481224
#>  [5,] -0.7071068 -0.4850713 -1.16599767 -0.6728772  0.2136234  0.5834600
#>  [6,] -0.7071068 -0.4850713 -1.16599767 -0.7230786  0.7218997  0.5834600
#>  [7,] -0.7071068  1.2126781 -0.05299989 -0.4950451  0.7218997  0.5834600
#>  [8,] -0.7071068  1.2126781 -0.05299989 -0.4950451  0.7218997  0.5834600
#>  [9,] -0.7071068  0.3638034  1.05999788  0.4255978 -1.1565128 -0.9481224
#> [10,] -0.7071068  0.3638034  1.05999788  0.4255978 -1.1565128 -0.9481224
#> [11,] -0.7071068  0.3638034  1.05999788  0.4255978 -1.1565128 -0.9481224
#> [12,] -0.7071068  1.2126781  1.05999788  1.9929029 -1.3112056 -0.9481224
#> [13,] -0.7071068  1.2126781  1.05999788  1.8227286 -0.8029293 -0.9481224
#> [14,]  1.4142136 -1.3339459 -1.16599767 -1.3161360  1.3848689  0.5834600
#> [15,] -0.7071068 -1.3339459 -1.16599767 -0.8992090  0.2357224 -0.9481224
#> [16,] -0.7071068 -0.4850713  1.05999788  0.7846655 -1.8415809 -0.9481224
#> [17,] -0.7071068  1.2126781  1.05999788  1.0569444  0.3020193 -0.9481224
#> [18,]  1.4142136 -1.3339459 -1.16599767 -1.2489172  1.0754833  0.5834600
#> [19,]  1.4142136 -0.4850713 -1.16599767 -1.1119269  0.3904152  2.1150424
#> [20,]  1.4142136  1.2126781  1.05999788  1.0654531  1.3848689  2.1150424
#> [21,]  1.4142136 -0.4850713 -1.16599767 -0.8915512  1.1417802  0.5834600
#>               hp        qsec         vs          wt
#>  [1,] -0.5433483 -0.95234706 -1.0488088 -0.64912701
#>  [2,] -0.8307192  0.26573990  0.9534626 -0.95889069
#>  [3,] -0.5433483  0.73597813  0.9534626 -0.03476237
#>  [4,]  0.5554227 -0.63507790 -1.0488088  0.19756039
#>  [5,] -1.3547484  1.05324729  0.9534626 -0.06057601
#>  [6,] -0.7969108  2.69624832  0.9534626 -0.10187783
#>  [7,] -0.3235941  0.09010876  0.9534626  0.19756039
#>  [8,] -0.3235941  0.43004001  0.9534626  0.19756039
#>  [9,]  0.6399436 -0.41978811 -1.0488088  0.84806413
#> [10,]  0.6399436 -0.30647770 -1.0488088  0.49699862
#> [11,]  0.6399436 -0.07985686 -1.0488088  0.54862590
#> [12,]  1.2315895 -0.18183624 -1.0488088  2.24613088
#> [13,]  1.4851520 -0.40845707 -1.0488088  2.16455978
#> [14,] -1.3040359  0.99659208  0.9534626 -1.45967531
#> [15,] -0.7631025  1.05891281  0.9534626 -0.80917158
#> [16,]  0.1328185 -0.72006071 -1.0488088  0.28016404
#> [17,]  1.7387146 -1.54722674 -1.0488088  0.61057864
#> [18,] -1.2871318  0.43004001  0.9534626 -1.35642075
#> [19,] -0.4926358 -0.70306415  0.9534626 -1.79215500
#> [20,]  2.0598938 -2.06278914 -1.0488088 -0.08122692
#> [21,] -0.5602525  0.26007438  0.9534626 -0.48391971
#> attr(,"center")
#>  [1]   0.3333333   2.5714286   6.0952381 225.7809524   3.5933333   3.6190476
#>  [7] 142.1428571  18.1409524   0.5238095   3.2486667
#> attr(,"scale")
#>  [1]   0.4714045   1.1780302   1.7969488 117.5265709   0.4525098   0.6529195
#>  [7]  59.1570031   1.7650628   0.4994328   0.9684802
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 29
#> 
#> $lambda.min
#> [1] 0.6788947
#> 
#> $null.dev
#> [1] 30.96626
#> 
#> $Bias
#> [1] 2.889533
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 7.681637