Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 35.301584 33.209988 29.927866 27.045571 24.833097 23.183358 22.004553
#>   [8] 21.218085 20.756758 20.563212 20.588571 20.791268 21.136030 21.592998
#>  [15] 22.136969 22.733925 22.420900 22.099348 21.786637 21.292724 20.873064
#>  [22] 20.521169 18.796811 16.683458 14.564734 13.653998 13.101011 11.604633
#>  [29] 11.433640 11.287053 10.640099 10.842416 12.117039 12.070879 12.123259
#>  [36] 12.720742 12.613432 11.342309 11.184305  9.511756  9.298421  8.636918
#>  [43]  7.755035  7.754809  8.278295  8.567859 10.798339 11.141168 11.207411
#>  [50] 11.914596 13.722737 13.723311 13.670078 13.673200 13.562834 13.563299
#>  [57] 13.563743 13.343354 13.331449 13.914571 13.914930 13.915275 13.871608
#>  [64] 13.741061 13.950173 14.002549 14.002864 30.357191 30.281124 30.282848
#>  [71] 30.284548 29.751925 29.513431 29.683520 29.176117 29.174028 29.171960
#>  [78] 29.169909 29.167877 29.178406 29.194568 29.153070 29.151061 29.463726
#>  [85] 29.462052 29.460391 29.439038 29.437535 29.503126 29.493053 29.538286
#>  [92] 29.536992 29.535706 29.534428 29.533158 29.531896 29.530642 29.529395
#>  [99] 29.528156 29.526925
#> 
#> $cvse
#>   [1]  9.795495  9.740448  9.132884  8.485192  7.867100  7.277025  6.722529
#>   [8]  6.217879  5.782074  5.436510  5.201502  5.091583  5.111049  5.252290
#>  [15]  5.498141  5.824448  5.787241  5.703671  5.661711  5.607543  5.555070
#>  [22]  5.505299  4.553937  3.687674  3.365709  3.336441  3.302060  2.951139
#>  [29]  2.921479  2.901197  2.953588  2.914250  2.933992  2.922234  2.939965
#>  [36]  2.878094  2.910805  2.837915  2.807317  2.399340  2.120444  2.148985
#>  [43]  2.158328  2.158231  2.203048  2.199156  3.298516  3.584961  3.575939
#>  [50]  4.220817  5.526854  5.527194  5.519292  5.534596  5.517075  5.517345
#>  [57]  5.517597  5.539878  5.541375  5.506004  5.506196  5.506374  5.510094
#>  [64]  5.413391  5.404471  5.399515  5.399638 20.369309 20.377903 20.379448
#>  [71] 20.380988 20.414063 20.423850 20.412134 20.088756 20.087114 20.085483
#>  [78] 20.083862 20.082253 20.079781 20.079095 20.079055 20.077485 20.065231
#>  [85] 20.063675 20.062130 20.061204 20.059679 20.055455 20.054056 20.052277
#>  [92] 20.050792 20.049318 20.047854 20.046400 20.044956 20.043523 20.042099
#>  [99] 20.040685 20.039281
#> 
#> $fold
#>  [1]  8  5  4  3  9  2  9  1  7  1  6  7 10 10  5  6  1  4  3  8  2
#> 
#> $lambda
#>   [1] 5.009354770 4.671741025 4.356881315 4.063242097 3.789393179 3.534000761
#>   [7] 3.295820938 3.073693637 2.866536973 2.673341975 2.493167673 2.325136517
#>  [13] 2.168430097 2.022285166 1.885989914 1.758880506 1.640337847 1.529784566
#>  [19] 1.426682206 1.330528601 1.240855426 1.157225923 1.079232769 1.006496093
#>  [25] 0.938661625 0.875398973 0.816400012 0.761377383 0.710063096 0.662207220
#>  [31] 0.617576670 0.575954069 0.537136692 0.500935477 0.467174102 0.435688132
#>  [37] 0.406324210 0.378939318 0.353400076 0.329582094 0.307369364 0.286653697
#>  [43] 0.267334198 0.249316767 0.232513652 0.216843010 0.202228517 0.188598991
#>  [49] 0.175888051 0.164033785 0.152978457 0.142668221 0.133052860 0.124085543
#>  [55] 0.115722592 0.107923277 0.100649609 0.093866162 0.087539897 0.081640001
#>  [61] 0.076137738 0.071006310 0.066220722 0.061757667 0.057595407 0.053713669
#>  [67] 0.050093548 0.046717410 0.043568813 0.040632421 0.037893932 0.035340008
#>  [73] 0.032958209 0.030736936 0.028665370 0.026733420 0.024931677 0.023251365
#>  [79] 0.021684301 0.020222852 0.018859899 0.017588805 0.016403378 0.015297846
#>  [85] 0.014266822 0.013305286 0.012408554 0.011572259 0.010792328 0.010064961
#>  [91] 0.009386616 0.008753990 0.008164000 0.007613774 0.007100631 0.006622072
#>  [97] 0.006175767 0.005759541 0.005371367 0.005009355
#> 
#> $fit
#> $beta
#>               5.0094     4.6717     4.3569    4.0632    3.7894    3.5340
#> (Intercept) 20.62857 22.2072958 23.6796194 25.052713 26.333265 27.527512
#> am           0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> wt           0.00000 -0.5024965 -0.9711264 -1.408172 -1.815763 -2.195883
#>                3.2958    3.0737   2.8665    2.6733    2.4932    2.3251
#> (Intercept) 28.641271 29.679966 30.64866 31.552061 32.394579 33.180314
#> am           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> wt          -2.550384 -2.880993 -3.18932 -3.476867 -3.745035 -3.995128
#>                2.1684    2.0223    1.8860    1.7589    1.6403     1.5298
#> (Intercept) 33.913092 34.596485 35.233818 35.828198 36.244820 35.8111804
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.0227704
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> wt          -4.228367 -4.445885 -4.648744 -4.837931 -4.970539 -4.9614707
#>                 1.4267     1.3305     1.2409     1.1572     1.0792     1.0065
#> (Intercept) 34.0893809 32.4837268 30.9862884 29.5897723 28.2873767 27.0731658
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.1131821  0.1974949  0.2761254  0.3494564  0.4178451  0.4816033
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -4.9254660 -4.8918901 -4.8605771 -4.8313745 -4.8041400 -4.7787496
#>                 0.9387     0.8754     0.8164     0.7614     0.7101     0.6622
#> (Intercept) 25.9403813 24.8839425 23.8987041 22.9798675 22.1229574 21.3238002
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.5410858  0.5965594  0.6482942  0.6965423  0.7415386  0.7835024
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -4.7550619 -4.7329707 -4.7123683 -4.6931545 -4.6752356 -4.6585243
#>                 0.6176     0.5760     0.5371     0.5009    0.4672     0.4357
#> (Intercept) 20.5785035 19.8834373 19.3569606 19.3569069 19.356904 18.1557336
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.3658131
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> qsec         0.8226379  0.8591358  0.8867811  0.8867839  0.886784  0.9202830
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> wt          -4.6429394 -4.6284049 -4.6173957 -4.6173945 -4.617394 -4.4802323
#>                 0.4063    0.3789    0.3534    0.3296    0.3074    0.2867
#> (Intercept) 16.0369849 12.912541  9.996275  8.763542  8.760795  8.758726
#> am           1.0105466  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.0000000 -1.905113 -1.893001 -1.887537 -1.887716 -1.887852
#> cyl          0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.0000000  1.224133  1.918659  2.212929  2.213012  2.213071
#> gear         0.0000000  3.752179  3.674156  3.640476  3.640971  3.641346
#> hp           0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.9794145  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -4.2384523 -1.611495 -1.409077 -1.323810 -1.323448 -1.323174
#>                0.2673    0.2493    0.2325    0.2168    0.2022      0.1886
#> (Intercept)  8.757164  8.755983  8.755090  8.754413  8.753901 -7.79869479
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> carb        -1.887955 -1.888034 -1.888093 -1.888139 -1.888173 -2.31614238
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  1.46479722
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> drat         2.213113  2.213145  2.213169  2.213186  2.213199  3.45448665
#> gear         3.641631  3.641848  3.642012  3.642136  3.642231  5.35768217
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000 -0.02076424
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> wt          -1.322966 -1.322809 -1.322690 -1.322600 -1.322532 -0.93452279
#>                  0.1759      0.1640       0.1530      0.1427       0.1331
#> (Intercept) -7.79454435 -7.79074819 -13.11915228 -13.1239645 -13.12841474
#> am           0.00000000  0.00000000  -1.46510130  -1.4653375  -1.46555542
#> carb        -2.31590278 -2.31568347  -2.36772135  -2.3679485  -2.36815894
#> cyl          1.46452518  1.46428068   1.69338473   1.6936760   1.69394602
#> disp         0.00000000  0.00000000   0.00000000   0.0000000   0.00000000
#> drat         3.45422159  3.45397519   4.16587294   4.1662549   4.16660743
#> gear         5.35715378  5.35667282   6.13353200   6.1341465   6.13471524
#> hp          -0.02076365 -0.02076318  -0.02230481  -0.0223057  -0.02230653
#> qsec         0.00000000  0.00000000   0.00000000   0.0000000   0.00000000
#> vs           0.00000000  0.00000000   0.00000000   0.0000000   0.00000000
#> wt          -0.93464514 -0.93475986  -1.07751845  -1.0774217  -1.07733265
#>                   0.1241     0.1157       0.1079       0.1006       0.0939
#> (Intercept) -13.13253011 -13.136336 -13.13985488 -13.14310909 -13.14611831
#> am           -1.46575634  -1.465942  -1.46611236  -1.46626979  -1.46641493
#> carb         -2.36835382  -2.368534  -2.36870144  -2.36885621  -2.36899953
#> cyl           1.69419612   1.694428   1.69464212   1.69484052   1.69502406
#> disp          0.00000000   0.000000   0.00000000   0.00000000   0.00000000
#> drat          4.16693283   4.167233   4.16751070   4.16776696   4.16800371
#> gear          6.13524143   6.135728   6.13617857   6.13659512   6.13698038
#> hp           -0.02230729  -0.022308  -0.02230865  -0.02230926  -0.02230981
#> qsec          0.00000000   0.000000   0.00000000   0.00000000   0.00000000
#> vs            0.00000000   0.000000   0.00000000   0.00000000   0.00000000
#> wt           -1.07725053  -1.077175  -1.07710479  -1.07704012  -1.07698031
#>                   0.0875       0.0816       0.0761       0.0710       0.0662
#> (Intercept) -13.14890097 -13.15147416 -13.15385367 -13.15605408 -15.32259608
#> am           -1.46654875  -1.46667216  -1.46678599  -1.46689101  -0.50333484
#> carb         -2.36913222  -2.36925507  -2.36936878  -2.36947403  -1.28045572
#> cyl           1.69519382   1.69535081   1.69549595   1.69563013   3.26335112
#> disp          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> drat          4.16822247   4.16842465   4.16861154   4.16878433   2.82211521
#> gear          6.13733669   6.13766620   6.13797091   6.13825269   5.12301307
#> hp           -0.02231032  -0.02231079  -0.02231122  -0.02231162  -0.04678047
#> qsec          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> vs            0.00000000   0.00000000   0.00000000   0.00000000   6.03001082
#> wt           -1.07692496  -1.07687372  -1.07682627  -1.07678232  -1.44869615
#>                   0.0618       0.0576       0.0537       0.0501      0.0467
#> (Intercept) -15.32235900 -15.32212595 -15.32189686 -15.32167166 -15.3214503
#> am           -0.50309552  -0.50286026  -0.50262901  -0.50240168  -0.5021782
#> carb         -1.28022921  -1.28000654  -1.27978766  -1.27957250  -1.2793610
#> cyl           3.26363868   3.26392135   3.26419921   3.26447235   3.2647409
#> disp          0.00000000   0.00000000   0.00000000   0.00000000   0.0000000
#> drat          2.82178186   2.82145417   2.82113205   2.82081540   2.8205041
#> gear          5.12274874   5.12248891   5.12223350   5.12198242   5.1217356
#> hp           -0.04678513  -0.04678971  -0.04679422  -0.04679864  -0.0468030
#> qsec          0.00000000   0.00000000   0.00000000   0.00000000   0.0000000
#> vs            6.03115118   6.03227216   6.03337409   6.03445730   6.0355221
#> wt           -1.44880478  -1.44891155  -1.44901651  -1.44911969  -1.4492211
#>                  0.0436      0.0406      0.0379      0.0353      0.0330
#> (Intercept) -9.91990624 -9.91659831 -9.91338025 -9.91024998 -9.90720544
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb        -0.72964472 -0.72955404 -0.72946908 -0.72938965 -0.72931556
#> cyl          3.07878907  3.07848217  3.07818037  3.07788361  3.07759182
#> disp         0.01054671  0.01054898  0.01055116  0.01055326  0.01055528
#> drat         2.07339155  2.07320266  2.07302220  2.07284992  2.07268556
#> gear         4.51107494  4.51084853  4.51063014  4.51041955  4.51021655
#> hp          -0.05728007 -0.05727939 -0.05727865 -0.05727785 -0.05727700
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           7.19066992  7.19028740  7.18990125  7.18951183  7.18911947
#> wt          -2.22098438 -2.22118963 -2.22138607 -2.22157396 -2.22175357
#>                  0.0307      0.0287      0.0267      0.0249       0.0233
#> (Intercept) -9.90424463 -9.90136559 -9.89856638 -9.89584513 -11.27580266
#> am           0.00000000  0.00000000  0.00000000  0.00000000   0.00000000
#> carb        -0.72924663 -0.72918267 -0.72912351 -0.72906899  -0.73507057
#> cyl          3.07730496  3.07702296  3.07674578  3.07647334   3.02903572
#> disp         0.01055722  0.01055909  0.01056088  0.01056260   0.01154257
#> drat         2.07252888  2.07237963  2.07223759  2.07210253   2.11813044
#> gear         4.51002091  4.50983243  4.50965092  4.50947617   4.54992072
#> hp          -0.05727611 -0.05727517 -0.05727418 -0.05727315  -0.05598192
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000   0.08668640
#> vs           7.18872450  7.18832724  7.18792797  7.18752699   6.86960144
#> wt          -2.22192514 -2.22208893 -2.22224516 -2.22239409  -2.36609512
#>                   0.0217       0.0202       0.0189       0.0176       0.0164
#> (Intercept) -11.27538265 -11.27498935 -11.27462195 -11.27427963 -11.27396159
#> am            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> carb         -0.73526820  -0.73546454  -0.73565954  -0.73585317  -0.73604540
#> cyl           3.02873287   3.02843600   3.02814500   3.02785978   3.02758025
#> disp          0.01154189   0.01154120   0.01154050   0.01153977   0.01153903
#> drat          2.11831614   2.11850136   2.11868603   2.11887009   2.11905349
#> gear          4.55000547   4.55009107   4.55017744   4.55026452   4.55035224
#> hp           -0.05597641  -0.05597096  -0.05596558  -0.05596027  -0.05595502
#> qsec          0.08670972   0.08673243   0.08675453   0.08677605   0.08679699
#> vs            6.86849368   6.86740200   6.86632626   6.86526631   6.86422201
#> wt           -2.36597564  -2.36585492  -2.36573307  -2.36561016  -2.36548630
#>                   0.0153       0.0143       0.0133       0.0124       0.0116
#> (Intercept) -11.27366707 -11.27339530 -11.27314556 -11.27291712 -11.27270928
#> am            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> carb         -0.73623620  -0.73642554  -0.73661340  -0.73679976  -0.73698458
#> cyl           3.02730630   3.02703786   3.02677482   3.02651710   3.02626461
#> disp          0.01153828   0.01153751   0.01153674   0.01153595   0.01153514
#> drat          2.11923618   2.11941811   2.11959922   2.11977947   2.11995881
#> gear          4.55044055   4.55052938   4.55061868   4.55070839   4.55079846
#> hp           -0.05594985  -0.05594474  -0.05593969  -0.05593472  -0.05592981
#> qsec          0.08681737   0.08683719   0.08685649   0.08687525   0.08689350
#> vs            6.86319324   6.86217983   6.86118166   6.86019857   6.85923042
#> wt           -2.36536158  -2.36523607  -2.36510987  -2.36498306  -2.36485571
#>                   0.0108       0.0101       0.0094       0.0088       0.0082
#> (Intercept) -11.27252136 -11.27235269 -11.27220262 -11.27207053 -11.27195578
#> am            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> carb         -0.73716784  -0.73734954  -0.73752964  -0.73770813  -0.73788500
#> cyl           3.02601725   3.02577495   3.02553761   3.02530516   3.02507750
#> disp          0.01153433   0.01153351   0.01153268   0.01153184   0.01153100
#> drat          2.12013721   2.12031461   2.12049099   2.12066630   2.12084052
#> gear          4.55088883   4.55097946   4.55107030   4.55116130   4.55125241
#> hp           -0.05592496  -0.05592018  -0.05591546  -0.05591081  -0.05590623
#> qsec          0.08691125   0.08692851   0.08694528   0.08696159   0.08697744
#> vs            6.85827706   6.85733834   6.85641412   6.85550425   6.85460858
#> wt           -2.36472789  -2.36459969  -2.36447116  -2.36434238  -2.36421341
#>                   0.0076       0.0071       0.0066       0.0062       0.0058
#> (Intercept) -11.27185778 -11.27177594 -11.27170969 -11.27165847 -11.27162174
#> am            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> carb         -0.73806023  -0.73823381  -0.73840572  -0.73857595  -0.73874450
#> cyl           3.02485455   3.02463624   3.02442247   3.02421318   3.02400829
#> disp          0.01153014   0.01152928   0.01152842   0.01152755   0.01152668
#> drat          2.12101360   2.12118551   2.12135623   2.12152573   2.12169398
#> gear          4.55134360   4.55143482   4.55152603   4.55161719   4.55170826
#> hp           -0.05590170  -0.05589724  -0.05589284  -0.05588851  -0.05588423
#> qsec          0.08699284   0.08700781   0.08702234   0.08703646   0.08705017
#> vs            6.85372695   6.85285922   6.85200523   6.85116484   6.85033789
#> wt           -2.36408431  -2.36395515  -2.36382597  -2.36369685  -2.36356783
#>                   0.0054       0.0050
#> (Intercept) -11.27159897 -11.27158965
#> am            0.00000000   0.00000000
#> carb         -0.73891135  -0.73907650
#> cyl           3.02380771   3.02361137
#> disp          0.01152580   0.01152492
#> drat          2.12186095   2.12202663
#> gear          4.55179921   4.55189001
#> hp           -0.05588002  -0.05587587
#> qsec          0.08706349   0.08707641
#> vs            6.84952424   6.84872373
#> wt           -2.36343895  -2.36331028
#> 
#> $iter
#>   [1]   0   5   3   3   2   2   2   2   2   2   2   2   2   2   2   2   2   4
#>  [19]   4   4   4   4   4   3   3   3   3   3   3   3   3   3   3   1   1  30
#>  [37]  33 127  31  14   1   1   1   1   1   1   1 116   1   1  64   1   1   1
#>  [55]   1   1   1   1   1   1   1   1 272   1   1   1   1   1 186   1   1   1
#>  [73]   1   1   1   1   1 136   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.009354770 4.671741025 4.356881315 4.063242097 3.789393179 3.534000761
#>   [7] 3.295820938 3.073693637 2.866536973 2.673341975 2.493167673 2.325136517
#>  [13] 2.168430097 2.022285166 1.885989914 1.758880506 1.640337847 1.529784566
#>  [19] 1.426682206 1.330528601 1.240855426 1.157225923 1.079232769 1.006496093
#>  [25] 0.938661625 0.875398973 0.816400012 0.761377383 0.710063096 0.662207220
#>  [31] 0.617576670 0.575954069 0.537136692 0.500935477 0.467174102 0.435688132
#>  [37] 0.406324210 0.378939318 0.353400076 0.329582094 0.307369364 0.286653697
#>  [43] 0.267334198 0.249316767 0.232513652 0.216843010 0.202228517 0.188598991
#>  [49] 0.175888051 0.164033785 0.152978457 0.142668221 0.133052860 0.124085543
#>  [55] 0.115722592 0.107923277 0.100649609 0.093866162 0.087539897 0.081640001
#>  [61] 0.076137738 0.071006310 0.066220722 0.061757667 0.057595407 0.053713669
#>  [67] 0.050093548 0.046717410 0.043568813 0.040632421 0.037893932 0.035340008
#>  [73] 0.032958209 0.030736936 0.028665370 0.026733420 0.024931677 0.023251365
#>  [79] 0.021684301 0.020222852 0.018859899 0.017588805 0.016403378 0.015297846
#>  [85] 0.014266822 0.013305286 0.012408554 0.011572259 0.010792328 0.010064961
#>  [91] 0.009386616 0.008753990 0.008164000 0.007613774 0.007100631 0.006622072
#>  [97] 0.006175767 0.005759541 0.005371367 0.005009355
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 35
#> 
#> $loss
#>   [1] 711.70286 610.54125 525.90453 455.41463 397.01844 348.94448 309.66519
#>   [8] 277.86436 252.40870 232.32326 216.77009 205.02972 196.48509 190.60763
#>  [15] 186.94520 185.11159 184.73652 182.03129 171.98438 163.61483 156.67876
#>  [22] 150.96633 146.29657 142.51473 139.48395 137.09015 135.23405 133.83039
#>  [29] 132.80604 132.09835 131.65372 131.42638 131.37451 131.37451 131.37451
#>  [36] 128.19837 124.04924  80.98783  79.31993  79.15760  79.15759  79.15759
#>  [43]  79.15759  79.15759  79.15758  79.15758  79.15758  63.70798  63.70796
#>  [50]  63.70795  61.18666  61.18665  61.18664  61.18663  61.18662  61.18661
#>  [57]  61.18660  61.18660  61.18659  61.18659  61.18658  61.18658  56.70477
#>  [64]  56.70475  56.70473  56.70471  56.70470  56.70468  55.48944  55.48943
#>  [71]  55.48942  55.48941  55.48940  55.48939  55.48938  55.48937  55.48936
#>  [78]  55.40597  55.40595  55.40593  55.40591  55.40589  55.40588  55.40586
#>  [85]  55.40584  55.40583  55.40581  55.40580  55.40578  55.40577  55.40575
#>  [92]  55.40574  55.40573  55.40571  55.40570  55.40569  55.40568  55.40567
#>  [99]  55.40566  55.40565
#> 
#> $linear.predictors
#>      5.0094   4.6717   4.3569   4.0632   3.7894   3.5340   3.2958   3.0737
#> 1  20.62857 20.89075 21.13527 21.36330 21.57597 21.77430 21.95926 22.13176
#> 2  20.62857 20.76262 20.88763 21.00422 21.11295 21.21435 21.30892 21.39711
#> 3  20.62857 21.04150 21.42661 21.78575 22.12070 22.43306 22.72438 22.99606
#> 4  20.62857 20.47871 20.33894 20.20860 20.08704 19.97368 19.86795 19.76935
#> 5  20.62857 20.41338 20.21270 20.02554 19.85099 19.68821 19.53640 19.39482
#> 6  20.62857 20.60433 20.58173 20.56064 20.54098 20.52265 20.50555 20.48960
#> 7  20.62857 20.62443 20.62057 20.61697 20.61361 20.61048 20.60756 20.60484
#> 8  20.62857 20.33298 20.05732 19.80023 19.56047 19.33687 19.12834 18.93386
#> 9  20.62857 20.30786 20.00876 19.72982 19.46968 19.22708 19.00082 18.78981
#> 10 20.62857 19.48175 18.41223 17.41479 16.48457 15.61704 14.80799 14.05346
#> 11 20.62857 19.52145 18.48895 17.52603 16.62801 15.79052 15.00947 14.28106
#> 12 20.62857 21.10180 21.54314 21.95473 22.33859 22.69657 23.03043 23.34178
#> 13 20.62857 21.39576 22.11125 22.77852 23.40081 23.98116 24.52240 25.02716
#> 14 20.62857 20.96864 21.28579 21.58157 21.85741 22.11466 22.35457 22.57832
#> 15 20.62857 20.27771 19.95049 19.64533 19.36074 19.09532 18.84780 18.61695
#> 16 20.62857 20.27520 19.94564 19.63829 19.35166 19.08434 18.83504 18.60255
#> 17 20.62857 21.23497 21.80049 22.32790 22.81976 23.27848 23.70628 24.10524
#> 18 20.62857 21.44702 22.21031 22.92215 23.58602 24.20514 24.78254 25.32102
#> 19 20.62857 20.81538 20.98960 21.15208 21.30360 21.44492 21.57671 21.69961
#> 20 20.62857 20.41338 20.21270 20.02554 19.85099 19.68821 19.53640 19.39482
#> 21 20.62857 20.81036 20.97989 21.13799 21.28545 21.42296 21.55120 21.67081
#>      2.8665   2.6733   2.4932   2.3251   2.1684   2.0223   1.8860    1.7589
#> 1  22.29264 22.44267 22.58259 22.71308 22.83477 22.94827 23.05411 23.152819
#> 2  21.47936 21.55607 21.62760 21.69432 21.75654 21.81456 21.86868 21.919147
#> 3  23.24943 23.48573 23.70610 23.91162 24.10328 24.28203 24.44873 24.604199
#> 4  19.67739 19.59164 19.51166 19.43707 19.36751 19.30264 19.24214 19.185716
#> 5  19.26278 19.13964 19.02480 18.91771 18.81782 18.72467 18.63780 18.556785
#> 6  20.47472 20.46085 20.44792 20.43585 20.42460 20.41411 20.40432 20.395199
#> 7  20.60230 20.59993 20.59772 20.59566 20.59374 20.59195 20.59027 20.588716
#> 8  18.75249 18.58335 18.42560 18.27848 18.14129 18.01333 17.89400 17.782716
#> 9  18.59303 18.40950 18.23835 18.07873 17.92987 17.79104 17.66157 17.540820
#> 10 13.34978 12.69353 12.08151 11.51074 10.97843 10.48200 10.01903  9.587262
#> 11 13.60174 12.96820 12.37737 11.82635 11.31247 10.83323 10.38628  9.969458
#> 12 23.63215 23.90295 24.15550 24.39103 24.61069 24.81554 25.00658 25.184750
#> 13 25.49790 25.93692 26.34635 26.72818 27.08428 27.41638 27.72610 28.014940
#> 14 22.78698 22.98158 23.16307 23.33232 23.49017 23.63738 23.77466 23.902699
#> 15 18.40167 18.20089 18.01365 17.83902 17.67616 17.52428 17.38264 17.250544
#> 16 18.38572 18.18351 17.99492 17.81904 17.65502 17.50206 17.35940 17.226354
#> 17 24.47732 24.82432 25.14794 25.44974 25.73120 25.99370 26.23850 26.466802
#> 18 25.82322 26.29156 26.72834 27.13568 27.51557 27.86986 28.20027 28.508409
#> 19 21.81424 21.92114 22.02083 22.11381 22.20052 22.28138 22.35680 22.427130
#> 20 19.26278 19.13964 19.02480 18.91771 18.81782 18.72467 18.63780 18.556785
#> 21 21.78235 21.88637 21.98338 22.07386 22.15823 22.23692 22.31031 22.378750
#>       1.6403    1.5298    1.4267    1.3305    1.2409    1.1572    1.0792
#> 1  23.222009 23.186928 23.047636 22.917741 22.796600 22.683623 22.578261
#> 2  21.954522 21.934505 21.855025 21.780906 21.711783 21.647318 21.587198
#> 3  24.713171 24.724326 24.768618 24.809922 24.848442 24.884367 24.917870
#> 4  19.146167 19.131274 19.072136 19.016988 18.965557 18.917592 18.872859
#> 5  18.499997 18.459413 18.298271 18.147998 18.007854 17.877154 17.755264
#> 6  20.388802 20.439497 20.640785 20.828495 21.003555 21.166815 21.319073
#> 7  20.587624 20.703990 21.166032 21.596906 21.998741 22.373494 22.722989
#> 8  17.704711 17.705654 17.709397 17.712887 17.716142 17.719178 17.722009
#> 9  17.456184 17.466689 17.508396 17.547290 17.583563 17.617392 17.648940
#> 10  9.284619  9.305932  9.390557  9.469474  9.543072  9.611710  9.675722
#> 11  9.677292  9.688780  9.734396  9.776935  9.816607  9.853606  9.888111
#> 12 25.309635 25.339285 25.457010 25.566794 25.669179 25.764664 25.853713
#> 13 28.217400 28.220113 28.230885 28.240930 28.250298 28.259035 28.267183
#> 14 23.992443 24.036791 24.212880 24.377091 24.530234 24.673056 24.806253
#> 15 17.157952 17.110025 16.919727 16.742265 16.576764 16.422417 16.278473
#> 16 17.133099 17.122561 17.080718 17.041697 17.005307 16.971369 16.939718
#> 17 26.626828 26.641095 26.697745 26.750573 26.799841 26.845788 26.888639
#> 18 28.724395 28.689295 28.549927 28.419961 28.298754 28.185716 28.080296
#> 19 22.476428 22.420848 22.200162 21.994362 21.802433 21.623439 21.456508
#> 20 18.499997 18.431178 18.157925 17.903105 17.665458 17.443829 17.237136
#> 21 22.426723 22.441821 22.501771 22.557678 22.609816 22.658440 22.703787
#>       1.0065    0.9387    0.8754    0.8164    0.7614    0.7101   0.6622
#> 1  22.480033 22.388391 22.302927 22.223222 22.148889 22.079566 22.01492
#> 2  21.531149 21.478859 21.430093 21.384613 21.342199 21.302643 21.26575
#> 3  24.949105 24.978245 25.005421 25.030765 25.054402 25.076445 25.09700
#> 4  18.831156 18.792249 18.755964 18.722125 18.690566 18.661135 18.63369
#> 5  17.641626 17.535609 17.436738 17.344530 17.258536 17.178338 17.10355
#> 6  21.461021 21.593450 21.716954 21.832134 21.939551 22.039729 22.13315
#> 7  23.048821 23.352801 23.636295 23.900682 24.147250 24.377200 24.59165
#> 8  17.724648 17.727111 17.729407 17.731549 17.733546 17.735409 17.73715
#> 9  17.678352 17.705792 17.731382 17.755248 17.777505 17.798262 17.81762
#> 10  9.735399  9.791075  9.842998  9.891422  9.936582  9.978698 10.01798
#> 11  9.920279  9.950290  9.978279 10.004381 10.028724 10.051426 10.07260
#> 12 25.936733 26.014186 26.086418 26.153782 26.216606 26.275196 26.32984
#> 13 28.274779 28.281865 28.288475 28.294638 28.300387 28.305748 28.31075
#> 14 24.930430 25.046281 25.154323 25.255084 25.349053 25.436690 25.51842
#> 15 16.144274 16.019076 15.902315 15.793424 15.691871 15.597163 15.50884
#> 16 16.910210 16.882681 16.857008 16.833065 16.810735 16.789910 16.77049
#> 17 26.928588 26.965858 27.000617 27.033032 27.063263 27.091457 27.11775
#> 18 27.982014 27.890323 27.804811 27.725063 27.650690 27.581329 27.51664
#> 19 21.300881 21.155690 21.020284 20.894004 20.776235 20.666404 20.56397
#> 20 17.044438 16.864663 16.697004 16.540645 16.394824 16.258831 16.13200
#> 21 22.746064 22.785505 22.822289 22.856593 22.888585 22.918421 22.94625
#>      0.6176   0.5760   0.5371   0.5009   0.4672   0.4357   0.4063   0.3789
#> 1  21.95462 21.89839 21.85580 21.85580 21.85580 21.93120 22.06395 20.85281
#> 2  21.23135 21.19926 21.17496 21.17496 21.17496 21.30410 21.53162 20.44187
#> 3  25.11617 25.13405 25.14760 25.14760 25.14760 25.25387 25.44123 26.99039
#> 4  18.60809 18.58422 18.56613 18.56613 18.56613 18.40695 18.12634 18.67133
#> 5  17.03379 16.96874 16.91947 16.91946 16.91946 16.73859 16.41964 14.72506
#> 6  22.22028 22.30154 22.36309 22.36310 22.36310 22.26945 22.10461 23.48741
#> 7  24.79165 24.97817 25.11945 25.11946 25.11947 25.11748 25.11445 23.83342
#> 8  17.73877 17.74028 17.74142 17.74142 17.74142 17.64145 17.46525 16.20095
#> 9  17.83567 17.85251 17.86526 17.86527 17.86527 17.78555 17.64510 16.12038
#> 10 10.05461 10.08877 10.11464 10.11465 10.11465 10.25440 10.50079 11.48028
#> 11 10.09234 10.11076 10.12471 10.12471 10.12471 10.24022 10.44386 11.88913
#> 12 26.38080 26.42832 26.46432 26.46432 26.46432 26.58295 26.79214 27.46532
#> 13 28.31541 28.31976 28.32305 28.32305 28.32305 28.32961 28.34119 27.54344
#> 14 25.59464 25.66573 25.71957 25.71957 25.71957 25.52682 25.18728 22.82092
#> 15 15.42647 15.34965 15.29146 15.29145 15.29145 15.13320 14.85411 14.92650
#> 16 16.75238 16.73549 16.72269 16.72269 16.72269 16.62007 16.43915 17.93298
#> 17 27.14227 27.16514 27.18246 27.18246 27.18246 27.24565 27.35706 27.89236
#> 18 27.45632 27.40006 27.35744 27.35744 27.35744 27.29574 27.18686 30.04000
#> 19 20.46845 20.37936 20.31188 20.31187 20.31187 20.37569 20.48794 20.21028
#> 20 16.01372 15.90341 15.81986 15.81985 15.81985 15.96325 16.21571 15.01292
#> 21 22.97220 22.99640 23.01473 23.01473 23.01473 23.18377 23.48174 24.66226
#>      0.3534   0.3296   0.3074   0.2867   0.2673   0.2493   0.2325   0.2168
#> 1  20.91188 20.93734 20.93713 20.93696 20.93684 20.93674 20.93667 20.93662
#> 2  20.55257 20.59977 20.59965 20.59955 20.59948 20.59943 20.59939 20.59936
#> 3  26.91767 26.88645 26.88666 26.88682 26.88694 26.88703 26.88710 26.88715
#> 4  18.42929 18.32672 18.32660 18.32652 18.32645 18.32640 18.32637 18.32634
#> 5  14.57523 14.51232 14.51190 14.51158 14.51134 14.51116 14.51102 14.51091
#> 6  23.49179 23.49313 23.49346 23.49371 23.49391 23.49405 23.49416 23.49424
#> 7  23.98945 24.05505 24.05539 24.05565 24.05584 24.05599 24.05610 24.05618
#> 8  15.97416 15.87824 15.87805 15.87790 15.87779 15.87770 15.87764 15.87759
#> 9  15.90371 15.81205 15.81187 15.81174 15.81164 15.81156 15.81150 15.81146
#> 10 11.55988 11.59327 11.59350 11.59367 11.59381 11.59391 11.59399 11.59404
#> 11 12.11249 12.20682 12.20704 12.20721 12.20734 12.20743 12.20751 12.20756
#> 12 27.52805 27.55428 27.55446 27.55460 27.55471 27.55479 27.55485 27.55490
#> 13 28.09022 28.32216 28.32203 28.32192 28.32184 28.32177 28.32173 28.32169
#> 14 22.75140 22.72208 22.72184 22.72165 22.72151 22.72140 22.72132 22.72126
#> 15 15.19248 15.30562 15.30534 15.30512 15.30496 15.30484 15.30474 15.30467
#> 16 17.72431 17.63567 17.63570 17.63572 17.63573 17.63574 17.63575 17.63576
#> 17 27.90146 27.90509 27.90518 27.90525 27.90530 27.90534 27.90537 27.90539
#> 18 29.68246 29.53067 29.53089 29.53107 29.53120 29.53130 29.53138 29.53144
#> 19 20.05145 19.98455 19.98450 19.98447 19.98444 19.98442 19.98441 19.98440
#> 20 14.98469 14.97339 14.97327 14.97318 14.97311 14.97306 14.97302 14.97298
#> 21 24.87535 24.96532 24.96554 24.96570 24.96583 24.96592 24.96599 24.96605
#>      0.2022   0.1886   0.1759   0.1640   0.1530   0.1427   0.1331   0.1241
#> 1  20.93658 21.89623 21.89630 21.89638 21.60957 21.60947 21.60937 21.60928
#> 2  20.59933 21.65793 21.65797 21.65801 21.33481 21.33473 21.33465 21.33458
#> 3  26.88720 26.37569 26.37563 26.37557 25.82011 25.82007 25.82004 25.82000
#> 4  18.32632 19.39358 19.39329 19.39304 19.60557 19.60586 19.60613 19.60638
#> 5  14.51083 13.39358 13.39378 13.39397 13.41867 13.41848 13.41830 13.41813
#> 6  23.49431 23.33748 23.33758 23.33767 24.00496 24.00498 24.00500 24.00501
#> 7  24.05625 23.48418 23.48424 23.48429 24.27015 24.27023 24.27030 24.27036
#> 8  15.87755 16.42624 16.42619 16.42614 16.48058 16.48063 16.48068 16.48073
#> 9  15.81143 16.37952 16.37946 16.37940 16.42670 16.42676 16.42681 16.42687
#> 10 11.59409 11.55846 11.55847 11.55849 11.21526 11.21519 11.21513 11.21507
#> 11 12.20760 12.11535 12.11533 12.11530 11.92396 11.92396 11.92396 11.92396
#> 12 27.55493 27.84300 27.84287 27.84276 27.50980 27.50986 27.50991 27.50996
#> 13 28.32166 29.30056 29.30052 29.30047 29.62568 29.62580 29.62590 29.62600
#> 14 22.72122 20.28127 20.28176 20.28221 20.28134 20.28088 20.28045 20.28005
#> 15 15.30461 14.93759 14.93762 14.93765 15.29399 15.29402 15.29405 15.29408
#> 16 17.63577 18.77328 18.77297 18.77268 18.87757 18.87787 18.87815 18.87840
#> 17 27.90541 28.09064 28.09055 28.09047 27.79534 27.79537 27.79540 27.79543
#> 18 29.53148 29.47974 29.47953 29.47933 29.67611 29.67634 29.67654 29.67673
#> 19 19.98439 20.16452 20.16464 20.16475 20.22978 20.22968 20.22959 20.22951
#> 20 14.97296 14.11557 14.11564 14.11571 14.11705 14.11699 14.11693 14.11687
#> 21 24.96609 24.19560 24.19567 24.19572 23.68298 23.68285 23.68272 23.68260
#>      0.1157   0.1079   0.1006   0.0939   0.0875   0.0816   0.0761   0.0710
#> 1  21.60920 21.60912 21.60905 21.60898 21.60892 21.60886 21.60881 21.60876
#> 2  21.33452 21.33446 21.33440 21.33435 21.33430 21.33426 21.33422 21.33418
#> 3  25.81997 25.81994 25.81992 25.81989 25.81987 25.81985 25.81983 25.81982
#> 4  19.60661 19.60682 19.60702 19.60720 19.60737 19.60752 19.60767 19.60780
#> 5  13.41798 13.41784 13.41771 13.41759 13.41747 13.41737 13.41728 13.41719
#> 6  24.00503 24.00504 24.00505 24.00506 24.00507 24.00507 24.00508 24.00509
#> 7  24.27041 24.27046 24.27051 24.27055 24.27059 24.27063 24.27066 24.27070
#> 8  16.48077 16.48081 16.48085 16.48089 16.48092 16.48095 16.48098 16.48100
#> 9  16.42691 16.42696 16.42700 16.42704 16.42707 16.42710 16.42713 16.42716
#> 10 11.21502 11.21497 11.21492 11.21488 11.21484 11.21480 11.21477 11.21474
#> 11 11.92396 11.92396 11.92395 11.92395 11.92395 11.92395 11.92395 11.92395
#> 12 27.51001 27.51006 27.51010 27.51014 27.51017 27.51020 27.51024 27.51026
#> 13 29.62608 29.62617 29.62624 29.62631 29.62637 29.62643 29.62649 29.62654
#> 14 20.27968 20.27933 20.27902 20.27873 20.27845 20.27820 20.27797 20.27776
#> 15 15.29410 15.29412 15.29414 15.29416 15.29418 15.29420 15.29421 15.29422
#> 16 18.87864 18.87887 18.87907 18.87926 18.87944 18.87960 18.87975 18.87989
#> 17 27.79546 27.79549 27.79551 27.79554 27.79556 27.79558 27.79559 27.79561
#> 18 29.67691 29.67707 29.67722 29.67736 29.67749 29.67761 29.67772 29.67782
#> 19 20.22943 20.22936 20.22930 20.22924 20.22918 20.22913 20.22908 20.22903
#> 20 14.11682 14.11678 14.11673 14.11669 14.11665 14.11662 14.11659 14.11656
#> 21 23.68249 23.68239 23.68229 23.68221 23.68213 23.68205 23.68198 23.68192
#>      0.0662   0.0618   0.0576   0.0537   0.0501   0.0467   0.0436   0.0406
#> 1  21.18922 21.18917 21.18913 21.18908 21.18904 21.18899 21.33246 21.33255
#> 2  20.81980 20.81973 20.81965 20.81958 20.81951 20.81944 20.76611 20.76615
#> 3  25.62267 25.62263 25.62260 25.62257 25.62254 25.62251 25.54245 25.54243
#> 4  19.31191 19.31187 19.31182 19.31178 19.31174 19.31170 19.44814 19.44813
#> 5  13.45736 13.45741 13.45746 13.45751 13.45756 13.45760 13.81492 13.81510
#> 6  24.58384 24.58389 24.58395 24.58401 24.58406 24.58411 24.73264 24.73264
#> 7  23.74711 23.74695 23.74678 23.74662 23.74646 23.74630 23.34589 23.34586
#> 8  17.15166 17.15181 17.15197 17.15212 17.15227 17.15241 16.73411 16.73396
#> 9  17.07922 17.07937 17.07952 17.07967 17.07981 17.07995 16.62306 16.62290
#> 10 11.58225 11.58231 11.58236 11.58242 11.58248 11.58253 12.03488 12.03492
#> 11 11.64407 11.64399 11.64392 11.64384 11.64377 11.64369 11.61708 11.61706
#> 12 27.70867 27.70870 27.70873 27.70876 27.70878 27.70881 27.52339 27.52327
#> 13 30.32942 30.32953 30.32963 30.32972 30.32982 30.32991 30.62568 30.62560
#> 14 20.18249 20.18250 20.18250 20.18251 20.18252 20.18253 20.29681 20.29706
#> 15 14.53371 14.53356 14.53341 14.53326 14.53312 14.53298 14.18795 14.18796
#> 16 18.52764 18.52757 18.52751 18.52745 18.52739 18.52733 18.82537 18.82538
#> 17 28.09257 28.09263 28.09269 28.09274 28.09280 28.09285 28.11511 28.11505
#> 18 29.47294 29.47289 29.47284 29.47280 29.47275 29.47270 29.66869 29.66870
#> 19 19.70309 19.70301 19.70293 19.70285 19.70277 19.70269 19.58915 19.58922
#> 20 14.79928 14.79942 14.79956 14.79969 14.79982 14.79995 14.82525 14.82521
#> 21 23.66107 23.66106 23.66104 23.66102 23.66100 23.66099 23.55086 23.55083
#>      0.0379   0.0353   0.0330   0.0307   0.0287   0.0267   0.0249   0.0233
#> 1  21.33264 21.33272 21.33280 21.33288 21.33296 21.33304 21.33312 21.33501
#> 2  20.76618 20.76622 20.76626 20.76629 20.76633 20.76637 20.76641 20.78020
#> 3  25.54242 25.54242 25.54241 25.54241 25.54241 25.54241 25.54241 25.49353
#> 4  19.44812 19.44811 19.44810 19.44809 19.44808 19.44807 19.44805 19.50274
#> 5  13.81528 13.81545 13.81561 13.81577 13.81592 13.81607 13.81621 13.83107
#> 6  24.73263 24.73263 24.73262 24.73261 24.73259 24.73258 24.73256 24.66368
#> 7  23.34584 23.34581 23.34579 23.34577 23.34574 23.34572 23.34571 23.58138
#> 8  16.73381 16.73366 16.73352 16.73338 16.73325 16.73311 16.73299 16.71053
#> 9  16.62274 16.62258 16.62243 16.62229 16.62214 16.62200 16.62187 16.62690
#> 10 12.03497 12.03501 12.03505 12.03509 12.03513 12.03517 12.03520 12.00487
#> 11 11.61705 11.61704 11.61703 11.61702 11.61702 11.61701 11.61701 11.57371
#> 12 27.52316 27.52305 27.52295 27.52285 27.52275 27.52266 27.52257 27.51249
#> 13 30.62552 30.62544 30.62535 30.62527 30.62519 30.62511 30.62503 30.62877
#> 14 20.29729 20.29753 20.29775 20.29797 20.29818 20.29838 20.29858 20.31990
#> 15 14.18796 14.18797 14.18798 14.18799 14.18800 14.18802 14.18803 14.14095
#> 16 18.82539 18.82540 18.82541 18.82542 18.82543 18.82543 18.82544 18.86050
#> 17 28.11499 28.11494 28.11488 28.11483 28.11477 28.11472 28.11467 28.09356
#> 18 29.66872 29.66874 29.66875 29.66876 29.66878 29.66879 29.66880 29.63160
#> 19 19.58929 19.58937 19.58944 19.58951 19.58958 19.58965 19.58972 19.57162
#> 20 14.82516 14.82512 14.82508 14.82503 14.82499 14.82495 14.82490 14.86273
#> 21 23.55081 23.55079 23.55077 23.55076 23.55075 23.55074 23.55073 23.47424
#>      0.0217   0.0202   0.0189   0.0176   0.0164   0.0153   0.0143   0.0133
#> 1  21.33508 21.33515 21.33522 21.33529 21.33535 21.33542 21.33548 21.33554
#> 2  20.78031 20.78043 20.78054 20.78065 20.78076 20.78087 20.78097 20.78108
#> 3  25.49364 25.49374 25.49385 25.49396 25.49406 25.49417 25.49427 25.49437
#> 4  19.50271 19.50268 19.50266 19.50263 19.50260 19.50257 19.50254 19.50252
#> 5  13.83103 13.83099 13.83095 13.83091 13.83087 13.83082 13.83078 13.83073
#> 6  24.66351 24.66333 24.66315 24.66298 24.66281 24.66264 24.66248 24.66231
#> 7  23.58150 23.58161 23.58172 23.58182 23.58193 23.58203 23.58213 23.58223
#> 8  16.71043 16.71032 16.71022 16.71012 16.71002 16.70992 16.70983 16.70974
#> 9  16.62681 16.62672 16.62663 16.62655 16.62646 16.62638 16.62630 16.62623
#> 10 12.00483 12.00479 12.00475 12.00471 12.00467 12.00463 12.00458 12.00454
#> 11 11.57379 11.57387 11.57394 11.57402 11.57410 11.57417 11.57425 11.57432
#> 12 27.51252 27.51255 27.51258 27.51261 27.51264 27.51267 27.51270 27.51274
#> 13 30.62859 30.62841 30.62824 30.62807 30.62790 30.62773 30.62757 30.62740
#> 14 20.31996 20.32002 20.32007 20.32013 20.32018 20.32022 20.32027 20.32031
#> 15 14.14104 14.14112 14.14121 14.14130 14.14138 14.14146 14.14155 14.14163
#> 16 18.86049 18.86047 18.86045 18.86043 18.86041 18.86039 18.86037 18.86035
#> 17 28.09354 28.09353 28.09351 28.09350 28.09348 28.09347 28.09346 28.09344
#> 18 29.63156 29.63152 29.63149 29.63145 29.63142 29.63138 29.63135 29.63131
#> 19 19.57169 19.57176 19.57183 19.57190 19.57197 19.57204 19.57210 19.57217
#> 20 14.86264 14.86255 14.86247 14.86238 14.86229 14.86221 14.86213 14.86204
#> 21 23.47434 23.47443 23.47452 23.47462 23.47471 23.47480 23.47489 23.47498
#>      0.0124   0.0116   0.0108   0.0101   0.0094   0.0088   0.0082   0.0076
#> 1  21.33560 21.33566 21.33572 21.33578 21.33583 21.33589 21.33594 21.33599
#> 2  20.78118 20.78128 20.78139 20.78148 20.78158 20.78168 20.78177 20.78187
#> 3  25.49448 25.49458 25.49468 25.49478 25.49487 25.49497 25.49507 25.49516
#> 4  19.50249 19.50246 19.50244 19.50241 19.50238 19.50236 19.50233 19.50230
#> 5  13.83069 13.83064 13.83059 13.83054 13.83049 13.83044 13.83039 13.83034
#> 6  24.66215 24.66199 24.66183 24.66168 24.66152 24.66137 24.66122 24.66108
#> 7  23.58233 23.58243 23.58252 23.58261 23.58270 23.58279 23.58288 23.58297
#> 8  16.70965 16.70956 16.70948 16.70940 16.70932 16.70924 16.70916 16.70909
#> 9  16.62615 16.62608 16.62601 16.62594 16.62587 16.62581 16.62574 16.62568
#> 10 12.00450 12.00446 12.00442 12.00438 12.00434 12.00431 12.00427 12.00423
#> 11 11.57440 11.57447 11.57454 11.57462 11.57469 11.57476 11.57483 11.57490
#> 12 27.51277 27.51280 27.51284 27.51287 27.51291 27.51294 27.51298 27.51301
#> 13 30.62724 30.62709 30.62693 30.62678 30.62663 30.62648 30.62634 30.62619
#> 14 20.32035 20.32039 20.32042 20.32046 20.32049 20.32052 20.32055 20.32057
#> 15 14.14171 14.14179 14.14187 14.14195 14.14203 14.14211 14.14218 14.14226
#> 16 18.86033 18.86031 18.86029 18.86027 18.86025 18.86023 18.86021 18.86019
#> 17 28.09343 28.09342 28.09341 28.09340 28.09339 28.09338 28.09338 28.09337
#> 18 29.63128 29.63125 29.63121 29.63118 29.63115 29.63111 29.63108 29.63105
#> 19 19.57223 19.57229 19.57235 19.57241 19.57247 19.57253 19.57259 19.57264
#> 20 14.86196 14.86188 14.86180 14.86172 14.86165 14.86157 14.86149 14.86142
#> 21 23.47507 23.47516 23.47525 23.47534 23.47542 23.47551 23.47560 23.47568
#>      0.0071   0.0066   0.0062   0.0058   0.0054   0.0050
#> 1  21.33604 21.33609 21.33614 21.33619 21.33624 21.33628
#> 2  20.78196 20.78205 20.78214 20.78223 20.78232 20.78240
#> 3  25.49526 25.49535 25.49544 25.49553 25.49562 25.49571
#> 4  19.50228 19.50225 19.50223 19.50220 19.50217 19.50215
#> 5  13.83029 13.83024 13.83019 13.83013 13.83008 13.83003
#> 6  24.66093 24.66079 24.66065 24.66051 24.66037 24.66024
#> 7  23.58305 23.58314 23.58322 23.58330 23.58338 23.58345
#> 8  16.70901 16.70894 16.70887 16.70881 16.70874 16.70868
#> 9  16.62562 16.62556 16.62550 16.62545 16.62540 16.62534
#> 10 12.00419 12.00415 12.00411 12.00408 12.00404 12.00400
#> 11 11.57497 11.57504 11.57510 11.57517 11.57524 11.57530
#> 12 27.51305 27.51308 27.51312 27.51316 27.51319 27.51323
#> 13 30.62605 30.62591 30.62578 30.62564 30.62551 30.62538
#> 14 20.32060 20.32062 20.32064 20.32066 20.32068 20.32069
#> 15 14.14233 14.14241 14.14248 14.14255 14.14262 14.14270
#> 16 18.86017 18.86015 18.86013 18.86012 18.86010 18.86008
#> 17 28.09336 28.09335 28.09335 28.09334 28.09334 28.09333
#> 18 29.63102 29.63099 29.63096 29.63093 29.63090 29.63087
#> 19 19.57270 19.57275 19.57281 19.57286 19.57291 19.57296
#> 20 14.86135 14.86128 14.86120 14.86113 14.86106 14.86100
#> 21 23.47577 23.47585 23.47593 23.47602 23.47610 23.47618
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 22.8 18.7 14.3 24.4 22.8 17.3 15.2 10.4 14.7 32.4 30.4 21.5 13.3
#> [16] 19.2 27.3 30.4 19.7 15.0 21.4
#> 
#> $X
#>               am        carb       cyl       disp        drat       gear
#>  [1,]  1.0488088  0.60507186  0.000000 -0.4835660  0.53997659  0.4082483
#>  [2,]  1.0488088  0.60507186  0.000000 -0.4835660  0.53997659  0.4082483
#>  [3,]  1.0488088 -1.12763391 -1.080123 -0.8920553  0.43239112  0.4082483
#>  [4,] -0.9534626 -0.55006532  1.080123  1.0875466 -1.07380544 -1.0206207
#>  [5,] -0.9534626  0.60507186  1.080123  1.0875466 -0.94470288 -1.0206207
#>  [6,] -0.9534626 -0.55006532 -1.080123 -0.5880450  0.08811762  0.4082483
#>  [7,] -0.9534626 -0.55006532 -1.080123 -0.6343929  0.58301078  0.4082483
#>  [8,] -0.9534626  0.02750327  1.080123  0.4261082 -1.24594219 -1.0206207
#>  [9,] -0.9534626  0.02750327  1.080123  0.4261082 -1.24594219 -1.0206207
#> [10,] -0.9534626  0.60507186  1.080123  1.8731030 -1.39656184 -1.0206207
#> [11,] -0.9534626  0.60507186  1.080123  1.7159917 -0.90166869 -1.0206207
#> [12,]  1.0488088 -1.12763391 -1.080123 -1.1222233  0.92728428  0.4082483
#> [13,]  1.0488088 -0.55006532 -1.080123 -1.1457900  2.75623724  0.4082483
#> [14,] -0.9534626 -1.12763391 -1.080123 -0.7970030  0.10963472 -1.0206207
#> [15,] -0.9534626  0.60507186  1.080123  1.0089910  0.17418600 -1.0206207
#> [16,] -0.9534626 -0.55006532  1.080123  1.4017692 -1.22442509 -1.0206207
#> [17,]  1.0488088 -1.12763391 -1.080123 -1.1198667  0.92728428  0.4082483
#> [18,]  1.0488088 -0.55006532 -1.080123 -0.9933921  0.26025437  1.8371173
#> [19,]  1.0488088  1.76020904  0.000000 -0.6013995 -0.06250203  1.8371173
#> [20,]  1.0488088  2.91534622  1.080123  0.6240684 -0.23463878  1.8371173
#> [21,]  1.0488088 -0.55006532 -1.080123 -0.7899330  0.99183556  0.4082483
#>               hp        qsec         vs          wt
#>  [1,] -0.5302035 -0.72365542 -0.8660254 -0.51771890
#>  [2,] -0.5302035 -0.41961156 -0.8660254 -0.26469483
#>  [3,] -0.7616003  0.44365584  1.1547005 -0.81539427
#>  [4,]  0.3545493 -0.41961156 -0.8660254  0.29592712
#>  [5,]  1.3073599 -1.06027541 -0.8660254  0.42491978
#>  [6,] -1.1835593  1.19833614  1.1547005  0.04786431
#>  [7,] -0.7343772  2.77284900  1.1547005  0.00817426
#>  [8,]  0.4226072 -0.10470898 -0.8660254  0.58367998
#>  [9,]  0.4226072  0.11246520 -0.8660254  0.63329255
#> [10,]  0.8990125  0.01473682 -0.8660254  2.26455360
#> [11,]  1.1031862 -0.20243737 -0.8660254  2.18616575
#> [12,] -1.1291130  0.91058034  1.1547005 -0.93446442
#> [13,] -1.3196751  0.39479165  1.1547005 -1.51493140
#> [14,] -0.7071540  1.20376550  1.1547005 -0.67151784
#> [15,]  1.3073599 -1.29373766 -0.8660254  0.69282762
#> [16,]  0.3545493 -0.40332349 -0.8660254  0.69778888
#> [17,] -1.1291130  0.60110713  1.1547005 -1.19741100
#> [18,] -0.4893687 -0.48476381  1.1547005 -1.61614103
#> [19,]  0.3545493 -1.24487347 -0.8660254 -0.36888121
#> [20,]  2.5324021 -1.73351539 -0.8660254  0.42491978
#> [21,] -0.5438150  0.43822649  1.1547005 -0.35895870
#> attr(,"center")
#>  [1]   0.4761905   2.9523810   6.0000000 221.5571429   3.6490476   3.7142857
#>  [7] 148.9523810  17.7928571   0.4285714   3.1417619
#> attr(,"scale")
#>  [1]   0.4994328   1.7313961   1.8516402 127.2983169   0.4647468   0.6998542
#>  [7]  73.4668543   1.8418395   0.4948717   1.0078093
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 44
#> 
#> $lambda.min
#> [1] 0.2493168
#> 
#> $null.dev
#> [1] 33.89061
#> 
#> $Bias
#> [1] 3.971238
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 13.47009