Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 35.88465 34.01109 31.00717 27.74309 25.20537 23.27910 21.86570 20.88072
#>   [9] 20.25189 19.91746 19.82476 19.92896 20.19195 20.58143 21.07009 21.20587
#>  [17] 20.69811 19.87102 19.33387 18.90090 18.55999 18.06718 17.39699 16.77181
#>  [25] 16.29458 15.19409 14.49030 13.87540 13.52251 13.29795 13.31472 13.64480
#>  [33] 14.00464 14.37396 17.16782 17.14252 17.61579 17.77677 17.81063 26.12011
#>  [41] 26.47279 26.48973 26.48602 28.12505 31.90465 31.89721 31.89796 31.64252
#>  [49] 34.44730 39.14696 42.50006 39.46616 39.48866 39.44234 39.44416 40.06835
#>  [57] 39.81633 40.45625 44.92213 43.22654 44.46268 44.46928 44.47469 44.47981
#>  [65] 44.48575 44.23032 46.94935 45.92638 46.51734 47.10203 47.10175 47.10158
#>  [73] 46.66926 46.66857 46.61207 46.54495 46.54459 46.54438 46.54169 46.54171
#>  [81] 46.54183 46.54205 45.77258 45.77182 45.73020 45.72952 45.72889 45.72829
#>  [89] 45.72774 45.72722 45.72673 45.72628 45.72586 45.72546 45.72508 45.72473
#>  [97] 45.72439 45.72408 45.72378 45.72349
#> 
#> $cvse
#>   [1] 12.347048 12.380349 11.756126 10.716762  9.763421  8.898892  8.132875
#>   [8]  7.480314  6.959106  6.586489  6.374105  6.323210  6.422628  6.650926
#>  [15]  6.981480  7.104909  6.960620  6.684579  6.507484  6.371740  6.250806
#>  [22]  6.142909  6.045913  5.961612  5.887502  5.727015  5.581040  5.422756
#>  [29]  5.273266  5.135272  5.044120  5.021696  5.001447  4.984554  5.105258
#>  [36]  5.101816  5.137959  5.171809  5.070053 10.248759 10.274653 10.269534
#>  [43] 10.267410 11.094223 11.465049 11.465832 11.466135 11.456800 12.626957
#>  [50] 14.989253 16.261523 14.617095 14.616269 14.623163 14.624181 14.821645
#>  [57] 14.719544 14.935333 15.451483 14.771401 15.272877 15.274845 15.276899
#>  [64] 15.278833 15.280954 15.191030 16.965556 16.506061 16.909234 17.354564
#>  [71] 17.355472 17.356367 17.229370 17.230034 17.237952 17.244090 17.244797
#>  [78] 17.245508 17.242152 17.242866 17.243580 17.244292 16.928017 16.928319
#>  [85] 16.931704 16.931994 16.932274 16.932545 16.932807 16.933060 16.933303
#>  [92] 16.933537 16.933762 16.933979 16.934186 16.934384 16.934574 16.934755
#>  [99] 16.934927 16.935091
#> 
#> $fold
#>  [1]  9  7  2  4  5  4  1 10  6  3  1  6  9  1  2 10  7  8  8  5  3
#> 
#> $lambda
#>   [1] 4.814730861 4.490234116 4.187607365 3.905376644 3.642167329 3.396697441
#>   [7] 3.167771402 2.954274211 2.755166017 2.569477049 2.396302895 2.234800100
#>  [13] 2.084182053 1.943715158 1.812715262 1.690544320 1.576607291 1.470349236
#>  [19] 1.371252619 1.278834782 1.192645598 1.112265276 1.037302319 0.967391615
#>  [25] 0.902192657 0.841387892 0.784681164 0.731796280 0.682475660 0.636479084
#>  [31] 0.593582524 0.553577049 0.516267809 0.481473086 0.449023412 0.418760736
#>  [37] 0.390537664 0.364216733 0.339669744 0.316777140 0.295427421 0.275516602
#>  [43] 0.256947705 0.239630290 0.223480010 0.208418205 0.194371516 0.181271526
#>  [49] 0.169054432 0.157660729 0.147034924 0.137125262 0.127883478 0.119264560
#>  [55] 0.111226528 0.103730232 0.096739161 0.090219266 0.084138789 0.078468116
#>  [61] 0.073179628 0.068247566 0.063647908 0.059358252 0.055357705 0.051626781
#>  [67] 0.048147309 0.044902341 0.041876074 0.039053766 0.036421673 0.033966974
#>  [73] 0.031677714 0.029542742 0.027551660 0.025694770 0.023963029 0.022348001
#>  [79] 0.020841821 0.019437152 0.018127153 0.016905443 0.015766073 0.014703492
#>  [85] 0.013712526 0.012788348 0.011926456 0.011122653 0.010373023 0.009673916
#>  [91] 0.009021927 0.008413879 0.007846812 0.007317963 0.006824757 0.006364791
#>  [97] 0.005935825 0.005535770 0.005162678 0.004814731
#> 
#> $fit
#> $beta
#>               4.8147     4.4902     4.1876     3.9054    3.6422    3.3967
#> (Intercept) 20.31429 22.0506703 23.6700284 25.1802471 26.588682 27.902193
#> am           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> cyl          0.00000 -0.2804929 -0.5420815 -0.7860399 -1.013556 -1.225739
#> disp         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> wt           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#>                3.1678    2.9543    2.7552    2.5695    2.3963    2.2348
#> (Intercept) 29.127178 30.269604 31.335033 32.328656 33.255312 34.119515
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.423621 -1.608167 -1.780275 -1.940783 -2.090474 -2.230076
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                2.0842    1.9437    1.8127    1.6905    1.5766    1.4703
#> (Intercept) 34.925473 35.677113 36.378094 37.031832 37.490060 37.490060
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.360269 -2.481687 -2.594923 -2.700527 -2.774548 -2.774548
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.3713    1.2788    1.1926     1.1123    1.0373    0.9674
#> (Intercept) 37.490060 37.490060 37.490060 37.9050361 38.780300 39.596620
#> am           0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> cyl         -2.774548 -2.774548 -2.774548 -2.6487318 -2.381864 -2.132954
#> disp         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000 -0.3713067 -1.157345 -1.890478
#>                0.9022    0.8414    0.7847    0.7318    0.6825    0.6365
#> (Intercept) 40.357539 41.032338 41.032222 41.032159 41.032124 41.032105
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.901062 -1.694969 -1.694854 -1.694792 -1.694757 -1.694739
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.573612 -3.180290 -3.180474 -3.180575 -3.180630 -3.180661
#>                0.5936      0.5536      0.5163     0.4815     0.4490     0.4188
#> (Intercept) 41.032095 41.01147860 40.96373480 40.9192663 40.8777956 40.8391592
#> am           0.000000  0.00000000  0.00000000  0.0000000  0.0000000  0.0000000
#> carb         0.000000 -0.02932647 -0.09747357 -0.1609867 -0.2202183 -0.2753243
#> cyl         -1.694728 -1.67948315 -1.64421906 -1.6113813 -1.5807574 -1.5522128
#> disp         0.000000  0.00000000  0.00000000  0.0000000  0.0000000  0.0000000
#> drat         0.000000  0.00000000  0.00000000  0.0000000  0.0000000  0.0000000
#> gear         0.000000  0.00000000  0.00000000  0.0000000  0.0000000  0.0000000
#> hp           0.000000  0.00000000  0.00000000  0.0000000  0.0000000  0.0000000
#> qsec         0.000000  0.00000000  0.00000000  0.0000000  0.0000000  0.0000000
#> vs           0.000000  0.00000000  0.00000000  0.0000000  0.0000000  0.0000000
#> wt          -3.180677 -3.17799140 -3.17149011 -3.1653851 -3.1596915 -3.1544808
#>                 0.3905     0.3642     0.3397     0.3168     0.2954     0.2755
#> (Intercept) 40.8030848 40.7694442 40.7380710 40.7088122 40.6815254 40.6687358
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb        -0.3268525 -0.3749053 -0.4197196 -0.4615135 -0.5004906 -0.5187607
#> cyl         -1.5255744 -1.5007335 -1.4775668 -1.4559614 -1.4358122 -1.4263683
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -3.1495233 -3.1448987 -3.1405858 -3.1365636 -3.1328125 -3.1310529
#>                0.2569    0.2396    0.2235    0.2084    0.1944    0.1813
#> (Intercept)  2.382788  2.379537  2.377392  2.375983  2.375064  2.374467
#> am           4.153628  4.153965  4.154183  4.154323  4.154413  4.154470
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         2.230202  2.230585  2.230837  2.231001  2.231109  2.231178
#> gear        -2.149864 -2.149969 -2.150036 -2.150078 -2.150104 -2.150120
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         1.402976  1.403036  1.403075  1.403100  1.403117  1.403128
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.839504 -2.839178 -2.838964 -2.838824 -2.838734 -2.838675
#>                0.1691    0.1577    0.1470    0.1371      0.1279      0.1193
#> (Intercept)  2.374082  2.373834  2.373676  2.373575  2.87791259  2.87560620
#> am           4.154507  4.154529  4.154544  4.154553  3.97312166  3.97316897
#> carb         0.000000  0.000000  0.000000  0.000000  0.00000000  0.00000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.00000000  0.00000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.00000000  0.00000000
#> drat         2.231222  2.231251  2.231269  2.231281  2.14649507  2.14676785
#> gear        -2.150130 -2.150137 -2.150140 -2.150143 -1.45474901 -1.45506697
#> hp           0.000000  0.000000  0.000000  0.000000 -0.01428285 -0.01427698
#> qsec         1.403135  1.403139  1.403142  1.403144  1.34375399  1.34388052
#> vs           0.000000  0.000000  0.000000  0.000000 -1.54675940 -1.54673029
#> wt          -2.838638 -2.838614 -2.838599 -2.838589 -2.42728024 -2.42749856
#>                  0.1112      0.1037      0.0967      0.0902      0.0841
#> (Intercept)  2.87369574  2.87211375  2.87080396  2.86971954  2.86882163
#> am           3.97320523  3.97323286  3.97325381  3.97326962  3.97328149
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         2.14700022  2.14719735  2.14736400  2.14750447  2.14762260
#> gear        -1.45533593 -1.45556277 -1.45575359 -1.45591376 -1.45604796
#> hp          -0.01427205 -0.01426793 -0.01426447 -0.01426159 -0.01425918
#> qsec         1.34398508  1.34407153  1.34414302  1.34420218  1.34425116
#> vs          -1.54670184 -1.54667524 -1.54665108 -1.54662963 -1.54661089
#> wt          -2.42768212 -2.42783636 -2.42796585 -2.42807450 -2.42816559
#>                  0.0785       0.0732       0.0682       0.0636      0.0594
#> (Intercept)  2.86807800  0.219654968  0.220500986  0.221320461 -0.75930405
#> am           3.97329036  4.162127944  4.162094692  4.162040226  4.62541564
#> carb         0.00000000  0.000000000  0.000000000  0.000000000  0.36872314
#> cyl          0.00000000  0.000000000  0.000000000  0.000000000  0.00000000
#> disp         0.00000000  0.007924786  0.007920023  0.007915423  0.01348476
#> drat         2.14772172  2.208341702  2.208478293  2.208614956  2.06901308
#> gear        -1.45616022 -1.188546558 -1.188944643 -1.189298466 -1.49963653
#> hp          -0.01425717 -0.019281277 -0.019275165 -0.019269777 -0.02579487
#> qsec         1.34429172  1.399974848  1.399980617  1.399980733  1.52087793
#> vs          -1.54659472 -1.294162550 -1.294187530 -1.294240829 -1.02473693
#> wt          -2.42824190 -2.679274734 -2.679209543 -2.679125991 -3.05085986
#>                  0.0554      0.0516      0.0481      0.0449      0.0419
#> (Intercept) -0.76060448 -0.76182827 -0.76297986 -0.76406342 -0.76508290
#> am           4.62581100  4.62618303  4.62653308  4.62686241  4.62717225
#> carb         0.36891335  0.36909227  0.36926057  0.36941885  0.36956771
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.01348953  0.01349401  0.01349823  0.01350220  0.01350593
#> drat         2.06892192  2.06883624  2.06875572  2.06868004  2.06860891
#> gear        -1.49984688 -1.50004499 -1.50023157 -1.50040726 -1.50057268
#> hp          -0.02579734 -0.02579966 -0.02580184 -0.02580389 -0.02580581
#> qsec         1.52098433  1.52108446  1.52117869  1.52126736  1.52135080
#> vs          -1.02443990 -1.02416037 -1.02389733 -1.02364982 -1.02341695
#> wt          -3.05119083 -3.05150227 -3.05179529 -3.05207098 -3.05233035
#>                  0.0391      0.0364      0.0340      0.0317      0.0295
#> (Intercept) -7.43305653 -7.43215957 -7.43118757 -7.43015199 -7.42906331
#> am           5.12466815  5.12488155  5.12507120  5.12523887  5.12538627
#> carb         0.35407211  0.35424398  0.35440606  0.35455879  0.35470258
#> cyl          0.51367800  0.51352289  0.51336755  0.51321256  0.51305847
#> disp         0.01199041  0.01199516  0.01199963  0.01200383  0.01200778
#> drat         2.32455118  2.32438950  2.32423304  2.32408184  2.32393589
#> gear        -1.51326766 -1.51346085 -1.51364265 -1.51381359 -1.51397418
#> hp          -0.02763703 -0.02763851 -0.02763986 -0.02764109 -0.02764221
#> qsec         1.66681332  1.66686277  1.66690660  1.66694524  1.66697909
#> vs          -0.56554910 -0.56539619 -0.56526427 -0.56515161 -0.56505662
#> wt          -2.98556161 -2.98588239 -2.98618462 -2.98646915 -2.98673678
#>                  0.0276      0.0257      0.0240      0.0223      0.0208
#> (Intercept) -7.42793113 -7.42676419 -7.42557043 -7.42435708 -7.42313066
#> am           5.12551497  5.12562646  5.12572215  5.12580332  5.12587121
#> carb         0.35483784  0.35496498  0.35508436  0.35519638  0.35530138
#> cyl          0.51290575  0.51275482  0.51260605  0.51245975  0.51231621
#> disp         0.01201148  0.01201494  0.01201819  0.01202123  0.01202407
#> drat         2.32379519  2.32365970  2.32352937  2.32340415  2.32328395
#> gear        -1.51412491 -1.51426627 -1.51439871 -1.51452268 -1.51463862
#> hp          -0.02764323 -0.02764415 -0.02764498 -0.02764574 -0.02764641
#> qsec         1.66700853  1.66703390  1.66705555  1.66707379  1.66708890
#> vs          -0.56497779 -0.56491371 -0.56486308 -0.56482466 -0.56479732
#> wt          -2.98698830 -2.98722447 -2.98744604 -2.98765371 -2.98784818
#>                  0.0194      0.0181      0.0169      0.0158      0.0147
#> (Intercept) -7.42189707 -7.42066162 -7.41942906 -7.41820362 -7.41698909
#> am           5.12592697  5.12597165  5.12600626  5.12603173  5.12604892
#> carb         0.35539973  0.35549174  0.35557775  0.35565808  0.35573302
#> cyl          0.51217567  0.51203834  0.51190437  0.51177392  0.51164709
#> disp         0.01202672  0.01202920  0.01203150  0.01203365  0.01203565
#> drat         2.32316870  2.32305830  2.32295265  2.32285163  2.32275514
#> gear        -1.51474693 -1.51484802 -1.51494228 -1.51503007 -1.51511176
#> hp          -0.02764702 -0.02764756 -0.02764804 -0.02764846 -0.02764883
#> qsec         1.66710116  1.66711083  1.66711813  1.66712330  1.66712653
#> vs          -0.56478000 -0.56477171 -0.56477153 -0.56477861 -0.56479216
#> wt          -2.98803013 -2.98820019 -2.98835898 -2.98850710 -2.98864513
#>                  0.0137      0.0128      0.0119      0.0111      0.0104
#> (Intercept) -7.41578880 -7.41460568 -7.41344230 -7.41230088 -7.41118331
#> am           5.12605864  5.12606164  5.12605861  5.12605020  5.12603700
#> carb         0.35580286  0.35586788  0.35592835  0.35598452  0.35603663
#> cyl          0.51152397  0.51140463  0.51128911  0.51117744  0.51106962
#> disp         0.01203750  0.01203922  0.01204081  0.01204229  0.01204365
#> drat         2.32266306  2.32257526  2.32249161  2.32241199  2.32233628
#> gear        -1.51518767 -1.51525813 -1.51532346 -1.51538396 -1.51543991
#> hp          -0.02764915 -0.02764943 -0.02764967 -0.02764987 -0.02765004
#> qsec         1.66712802  1.66712794  1.66712646  1.66712374  1.66711991
#> vs          -0.56481146 -0.56483583 -0.56486464 -0.56489732 -0.56493336
#> wt          -2.98877362 -2.98889309 -2.98900405 -2.98910699 -2.98920236
#>                  0.0097      0.0090      0.0084      0.0078      0.0073
#> (Intercept) -7.41009123 -7.40902599 -7.40798869 -7.40698023 -7.40600131
#> am           5.12601956  5.12599840  5.12597398  5.12594674  5.12591707
#> carb         0.35608492  0.35612962  0.35617093  0.35620906  0.35624420
#> cyl          0.51096564  0.51086550  0.51076914  0.51067654  0.51058764
#> disp         0.01204491  0.01204606  0.01204713  0.01204811  0.01204901
#> drat         2.32226433  2.32219602  2.32213122  2.32206980  2.32201163
#> gear        -1.51549158 -1.51553922 -1.51558309 -1.51562343 -1.51566044
#> hp          -0.02765017 -0.02765028 -0.02765036 -0.02765042 -0.02765046
#> qsec         1.66711510  1.66710944  1.66710303  1.66709598  1.66708837
#> vs          -0.56497225 -0.56501357 -0.56505689 -0.56510185 -0.56514812
#> wt          -2.98929061 -2.98937217 -2.98944743 -2.98951678 -2.98958058
#>                  0.0068      0.0064      0.0059      0.0055      0.0052
#> (Intercept) -7.40505244 -7.40413396 -7.40324608 -7.40238885 -7.40156222
#> am           5.12588533  5.12585187  5.12581698  5.12578095  5.12574403
#> carb         0.35627654  0.35630624  0.35633349  0.35635843  0.35638122
#> cyl          0.51050238  0.51042068  0.51034249  0.51026771  0.51019627
#> disp         0.01204983  0.01205058  0.01205126  0.01205189  0.01205245
#> drat         2.32195658  2.32190453  2.32185536  2.32180895  2.32176517
#> gear        -1.51569435 -1.51572535 -1.51575365 -1.51577940 -1.51580280
#> hp          -0.02765048 -0.02765048 -0.02765047 -0.02765045 -0.02765041
#> qsec         1.66708031  1.66707186  1.66706309  1.66705408  1.66704488
#> vs          -0.56519538 -0.56524336 -0.56529180 -0.56534048 -0.56538919
#> wt          -2.98963918 -2.98969291 -2.98974208 -2.98978700 -2.98982794
#>                  0.0048
#> (Intercept) -7.40076603
#> am           5.12570644
#> carb         0.35640199
#> cyl          0.51012808
#> disp         0.01205296
#> drat         2.32172392
#> gear        -1.51582400
#> hp          -0.02765036
#> qsec         1.66703554
#> vs          -0.56543775
#> wt          -2.98986516
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1  24  28  28  27  19   1   1   1   1   1   8   8   8   8   7
#>  [37]   7   7   7   7   7   5 154   1   1   1   1   1   1   1   1   1  77   1
#>  [55]   1   1   1   1   1   1  52   1   1  88   1   1   1   1   1  90   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.814730861 4.490234116 4.187607365 3.905376644 3.642167329 3.396697441
#>   [7] 3.167771402 2.954274211 2.755166017 2.569477049 2.396302895 2.234800100
#>  [13] 2.084182053 1.943715158 1.812715262 1.690544320 1.576607291 1.470349236
#>  [19] 1.371252619 1.278834782 1.192645598 1.112265276 1.037302319 0.967391615
#>  [25] 0.902192657 0.841387892 0.784681164 0.731796280 0.682475660 0.636479084
#>  [31] 0.593582524 0.553577049 0.516267809 0.481473086 0.449023412 0.418760736
#>  [37] 0.390537664 0.364216733 0.339669744 0.316777140 0.295427421 0.275516602
#>  [43] 0.256947705 0.239630290 0.223480010 0.208418205 0.194371516 0.181271526
#>  [49] 0.169054432 0.157660729 0.147034924 0.137125262 0.127883478 0.119264560
#>  [55] 0.111226528 0.103730232 0.096739161 0.090219266 0.084138789 0.078468116
#>  [61] 0.073179628 0.068247566 0.063647908 0.059358252 0.055357705 0.051626781
#>  [67] 0.048147309 0.044902341 0.041876074 0.039053766 0.036421673 0.033966974
#>  [73] 0.031677714 0.029542742 0.027551660 0.025694770 0.023963029 0.022348001
#>  [79] 0.020841821 0.019437152 0.018127153 0.016905443 0.015766073 0.014703492
#>  [85] 0.013712526 0.012788348 0.011926456 0.011122653 0.010373023 0.009673916
#>  [91] 0.009021927 0.008413879 0.007846812 0.007317963 0.006824757 0.006364791
#>  [97] 0.005935825 0.005535770 0.005162678 0.004814731
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 21
#> 
#> $loss
#>   [1] 692.40571 598.95209 520.76423 455.64529 401.69858 357.28760 321.00119
#>   [8] 291.62341 268.10734 249.55230 235.18420 224.33838 216.44481 211.01518
#>  [15] 207.63181 205.93791 205.59142 205.59142 205.59142 205.59142 205.59142
#>  [22] 192.16221 169.22599 154.55825 146.73248 144.50720 144.50720 144.50720
#>  [29] 144.50720 144.50720 144.50720 143.38147 141.01849 139.13457 137.65446
#>  [36] 136.51734 135.66324 135.04894 134.63450 134.38581 134.27374 134.26103
#>  [43] 103.78750 103.78750 103.78750 103.78750 103.78750 103.78749 103.78749
#>  [50] 103.78749 103.78749 103.78749  98.90339  98.90338  98.90338  98.90337
#>  [57]  98.90337  98.90337  98.90337  98.90337  97.56201  97.56200  97.56199
#>  [64]  96.75049  96.75048  96.75047  96.75046  96.75045  96.75044  96.00944
#>  [71]  96.00943  96.00942  96.00941  96.00940  96.00939  96.00938  96.00938
#>  [78]  96.00937  96.00937  96.00936  96.00936  96.00936  96.00935  96.00935
#>  [85]  96.00935  96.00935  96.00934  96.00934  96.00934  96.00934  96.00934
#>  [92]  96.00934  96.00934  96.00934  96.00933  96.00933  96.00933  96.00933
#>  [99]  96.00933  96.00933
#> 
#> $linear.predictors
#>      4.8147   4.4902   4.1876   3.9054   3.6422   3.3967   3.1678   2.9543
#> 1  20.31429 20.36771 20.41754 20.46401 20.50734 20.54776 20.58545 20.62060
#> 2  20.31429 20.92870 21.50170 22.03609 22.53446 22.99924 23.43269 23.83694
#> 3  20.31429 20.36771 20.41754 20.46401 20.50734 20.54776 20.58545 20.62060
#> 4  20.31429 19.80673 19.33338 18.89193 18.48023 18.09628 17.73821 17.40427
#> 5  20.31429 20.36771 20.41754 20.46401 20.50734 20.54776 20.58545 20.62060
#> 6  20.31429 19.80673 19.33338 18.89193 18.48023 18.09628 17.73821 17.40427
#> 7  20.31429 20.92870 21.50170 22.03609 22.53446 22.99924 23.43269 23.83694
#> 8  20.31429 20.92870 21.50170 22.03609 22.53446 22.99924 23.43269 23.83694
#> 9  20.31429 20.36771 20.41754 20.46401 20.50734 20.54776 20.58545 20.62060
#> 10 20.31429 20.36771 20.41754 20.46401 20.50734 20.54776 20.58545 20.62060
#> 11 20.31429 19.80673 19.33338 18.89193 18.48023 18.09628 17.73821 17.40427
#> 12 20.31429 19.80673 19.33338 18.89193 18.48023 18.09628 17.73821 17.40427
#> 13 20.31429 19.80673 19.33338 18.89193 18.48023 18.09628 17.73821 17.40427
#> 14 20.31429 20.92870 21.50170 22.03609 22.53446 22.99924 23.43269 23.83694
#> 15 20.31429 20.92870 21.50170 22.03609 22.53446 22.99924 23.43269 23.83694
#> 16 20.31429 20.92870 21.50170 22.03609 22.53446 22.99924 23.43269 23.83694
#> 17 20.31429 19.80673 19.33338 18.89193 18.48023 18.09628 17.73821 17.40427
#> 18 20.31429 19.80673 19.33338 18.89193 18.48023 18.09628 17.73821 17.40427
#> 19 20.31429 19.80673 19.33338 18.89193 18.48023 18.09628 17.73821 17.40427
#> 20 20.31429 19.80673 19.33338 18.89193 18.48023 18.09628 17.73821 17.40427
#> 21 20.31429 20.92870 21.50170 22.03609 22.53446 22.99924 23.43269 23.83694
#>      2.7552   2.5695   2.3963   2.2348   2.0842   1.9437   1.8127   1.6905
#> 1  20.65339 20.68396 20.71247 20.73906 20.76386 20.78699 20.80856 20.82867
#> 2  24.21393 24.56552 24.89342 25.19921 25.48440 25.75036 25.99840 26.22973
#> 3  20.65339 20.68396 20.71247 20.73906 20.76386 20.78699 20.80856 20.82867
#> 4  17.09284 16.80239 16.53152 16.27891 16.04332 15.82361 15.61871 15.42762
#> 5  20.65339 20.68396 20.71247 20.73906 20.76386 20.78699 20.80856 20.82867
#> 6  17.09284 16.80239 16.53152 16.27891 16.04332 15.82361 15.61871 15.42762
#> 7  24.21393 24.56552 24.89342 25.19921 25.48440 25.75036 25.99840 26.22973
#> 8  24.21393 24.56552 24.89342 25.19921 25.48440 25.75036 25.99840 26.22973
#> 9  20.65339 20.68396 20.71247 20.73906 20.76386 20.78699 20.80856 20.82867
#> 10 20.65339 20.68396 20.71247 20.73906 20.76386 20.78699 20.80856 20.82867
#> 11 17.09284 16.80239 16.53152 16.27891 16.04332 15.82361 15.61871 15.42762
#> 12 17.09284 16.80239 16.53152 16.27891 16.04332 15.82361 15.61871 15.42762
#> 13 17.09284 16.80239 16.53152 16.27891 16.04332 15.82361 15.61871 15.42762
#> 14 24.21393 24.56552 24.89342 25.19921 25.48440 25.75036 25.99840 26.22973
#> 15 24.21393 24.56552 24.89342 25.19921 25.48440 25.75036 25.99840 26.22973
#> 16 24.21393 24.56552 24.89342 25.19921 25.48440 25.75036 25.99840 26.22973
#> 17 17.09284 16.80239 16.53152 16.27891 16.04332 15.82361 15.61871 15.42762
#> 18 17.09284 16.80239 16.53152 16.27891 16.04332 15.82361 15.61871 15.42762
#> 19 17.09284 16.80239 16.53152 16.27891 16.04332 15.82361 15.61871 15.42762
#> 20 17.09284 16.80239 16.53152 16.27891 16.04332 15.82361 15.61871 15.42762
#> 21 24.21393 24.56552 24.89342 25.19921 25.48440 25.75036 25.99840 26.22973
#>      1.5766   1.4703   1.3713   1.2788   1.1926   1.1123   1.0373   0.9674
#> 1  20.84277 20.84277 20.84277 20.84277 20.84277 21.03982 21.45687 21.84585
#> 2  26.39187 26.39187 26.39187 26.39187 26.39187 26.44868 26.56780 26.67890
#> 3  20.84277 20.84277 20.84277 20.84277 20.84277 20.81889 20.76825 20.72101
#> 4  15.29367 15.29367 15.29367 15.29367 15.29367 15.43789 15.74412 16.02975
#> 5  20.84277 20.84277 20.84277 20.84277 20.84277 20.72792 20.48470 20.25785
#> 6  15.29367 15.29367 15.29367 15.29367 15.29367 15.38962 15.59366 15.78399
#> 7  26.39187 26.39187 26.39187 26.39187 26.39187 26.12564 25.56091 25.03418
#> 8  26.39187 26.39187 26.39187 26.39187 26.39187 26.14049 25.60721 25.10980
#> 9  20.84277 20.84277 20.84277 20.84277 20.84277 20.73535 20.50785 20.29566
#> 10 20.84277 20.84277 20.84277 20.84277 20.84277 20.73535 20.50785 20.29566
#> 11 15.29367 15.29367 15.29367 15.29367 15.29367 15.20396 15.01499 14.83875
#> 12 15.29367 15.29367 15.29367 15.29367 15.29367 15.33021 15.40849 15.48151
#> 13 15.29367 15.29367 15.29367 15.29367 15.29367 14.73055 13.53938 12.42839
#> 14 26.39187 26.39187 26.39187 26.39187 26.39187 26.49323 26.70668 26.90575
#> 15 26.39187 26.39187 26.39187 26.39187 26.39187 26.71045 27.38373 28.01168
#> 16 26.39187 26.39187 26.39187 26.39187 26.39187 26.62876 27.12911 27.59578
#> 17 15.29367 15.29367 15.29367 15.29367 15.29367 15.40818 15.65153 15.87851
#> 18 15.29367 15.29367 15.29367 15.29367 15.29367 15.28936 15.28118 15.27356
#> 19 15.29367 15.29367 15.29367 15.29367 15.29367 15.53814 16.05660 16.54018
#> 20 15.29367 15.29367 15.29367 15.29367 15.29367 15.38962 15.59366 15.78399
#> 21 26.39187 26.39187 26.39187 26.39187 26.39187 26.27788 26.03542 25.80928
#>      0.9022   0.8414   0.7847   0.7318   0.6825   0.6365   0.5936   0.5536
#> 1  22.20830 22.53017 22.53025 22.53030 22.53033 22.53034 22.53035 22.49094
#> 2  26.78251 26.87419 26.87411 26.87406 26.87403 26.87402 26.87401 26.89128
#> 3  20.67701 20.63789 20.63787 20.63786 20.63785 20.63785 20.63785 20.68801
#> 4  16.29582 16.53239 16.53256 16.53265 16.53270 16.53272 16.53274 16.58467
#> 5  20.04647 19.85872 19.85866 19.85862 19.85860 19.85859 19.85858 19.90940
#> 6  15.96125 16.11895 16.11910 16.11917 16.11922 16.11924 16.11925 16.11288
#> 7  24.54347 24.10734 24.10709 24.10696 24.10688 24.10684 24.10682 24.09710
#> 8  24.64641 24.23455 24.23431 24.23418 24.23411 24.23407 24.23405 24.22422
#> 9  20.09794 19.92233 19.92227 19.92223 19.92221 19.92220 19.92220 19.88498
#> 10 20.09794 19.92233 19.92227 19.92223 19.92221 19.92220 19.92220 19.88498
#> 11 14.67444 14.52881 14.52886 14.52889 14.52890 14.52891 14.52891 14.55321
#> 12 15.54947 15.61011 15.61022 15.61028 15.61031 15.61033 15.61034 15.63373
#> 13 11.39309 10.47394 10.47375 10.47365 10.47360 10.47357 10.47355 10.47194
#> 14 27.09135 27.25583 27.25576 27.25573 27.25571 27.25570 27.25569 27.27264
#> 15 28.59691 29.11630 29.11634 29.11636 29.11638 29.11638 29.11639 29.10244
#> 16 28.03071 28.41663 28.41664 28.41664 28.41664 28.41664 28.41664 28.43261
#> 17 16.08993 16.27797 16.27812 16.27820 16.27825 16.27827 16.27829 16.33043
#> 18 15.26637 15.26028 15.26037 15.26042 15.26045 15.26046 15.26047 15.25482
#> 19 16.99069 17.39107 17.39129 17.39140 17.39147 17.39150 17.39152 17.38407
#> 20 15.96125 16.11895 16.11910 16.11917 16.11922 16.11924 16.11925 15.99557
#> 21 25.59865 25.41126 25.41109 25.41099 25.41094 25.41091 25.41090 25.40008
#>      0.5163   0.4815   0.4490   0.4188   0.3905   0.3642   0.3397   0.3168
#> 1  22.39922 22.31372 22.23399 22.15985 22.09048 22.02579 21.96546 21.90919
#> 2  26.93153 26.96906 27.00406 27.03659 27.06704 27.09544 27.12193 27.14663
#> 3  20.80461 20.91328 21.01462 21.10890 21.19707 21.27929 21.35597 21.42748
#> 4  16.70511 16.81732 16.92196 17.01939 17.11042 17.19531 17.27448 17.34831
#> 5  20.02759 20.13776 20.24050 20.33605 20.42544 20.50879 20.58652 20.65902
#> 6  16.09787 16.08384 16.07076 16.05866 16.04728 16.03667 16.02677 16.01753
#> 7  24.07486 24.05419 24.03491 24.01687 24.00010 23.98447 23.96990 23.95630
#> 8  24.20172 24.18080 24.16130 24.14304 24.12608 24.11027 24.09552 24.08176
#> 9  19.79860 19.71811 19.64304 19.57317 19.50787 19.44697 19.39018 19.33721
#> 10 19.79860 19.71811 19.64304 19.57317 19.50787 19.44697 19.39018 19.33721
#> 11 14.60960 14.66214 14.71114 14.75675 14.79937 14.83912 14.87619 14.91077
#> 12 15.68790 15.73837 15.78543 15.82927 15.87021 15.90839 15.94399 15.97720
#> 13 10.46847 10.46529 10.46231 10.45946 10.45688 10.45447 10.45223 10.45013
#> 14 27.31211 27.34891 27.38323 27.41513 27.44498 27.47283 27.49880 27.52301
#> 15 29.06995 29.03967 29.01143 28.98517 28.96060 28.93769 28.91632 28.89639
#> 16 28.46970 28.50427 28.53651 28.56651 28.59456 28.62072 28.64511 28.66786
#> 17 16.45139 16.56409 16.66919 16.76704 16.85846 16.94372 17.02324 17.09739
#> 18 15.24157 15.22919 15.21765 15.20695 15.19691 15.18754 15.17881 15.17066
#> 19 17.36646 17.35000 17.33464 17.32046 17.30709 17.29463 17.28300 17.27216
#> 20 15.70797 15.43990 15.18989 14.95737 14.73987 14.53705 14.34789 14.17148
#> 21 25.37517 25.35200 25.33039 25.31020 25.29141 25.27388 25.25754 25.24229
#>      0.2954   0.2755   0.2569   0.2396   0.2235   0.2084   0.1944   0.1813
#> 1  21.85672 21.83212 22.28824 22.28823 22.28822 22.28821 22.28821 22.28821
#> 2  27.16966 27.18046 26.04498 26.04498 26.04498 26.04498 26.04498 26.04498
#> 3  21.49417 21.52543 20.94707 20.94689 20.94677 20.94669 20.94663 20.94660
#> 4  17.41717 17.44945 17.06910 17.06887 17.06872 17.06862 17.06855 17.06851
#> 5  20.72663 20.75832 20.63205 20.63187 20.63175 20.63167 20.63162 20.63159
#> 6  16.00892 16.00489 15.17826 15.17803 15.17787 15.17777 15.17771 15.17766
#> 7  23.94362 23.93768 21.01429 21.01426 21.01424 21.01423 21.01422 21.01422
#> 8  24.06894 24.06292 25.70945 25.70966 25.70981 25.70990 25.70996 25.71001
#> 9  19.28781 19.26466 18.43230 18.43234 18.43236 18.43238 18.43239 18.43240
#> 10 19.28781 19.26466 19.27409 19.27416 19.27421 19.27424 19.27426 19.27428
#> 11 14.94301 14.95812 15.63493 15.63489 15.63487 15.63486 15.63485 15.63484
#> 12 16.00817 16.02268 16.88095 16.88082 16.88073 16.88068 16.88064 16.88062
#> 13 10.44818 10.44727 12.39945 12.39990 12.40019 12.40038 12.40050 12.40058
#> 14 27.54560 27.55619 28.10523 28.10533 28.10539 28.10543 28.10546 28.10548
#> 15 28.87780 28.86909 30.32918 30.32936 30.32948 30.32955 30.32960 30.32963
#> 16 28.68907 28.69902 30.05715 30.05721 30.05725 30.05728 30.05729 30.05730
#> 17 17.16655 17.19896 15.76171 15.76135 15.76111 15.76096 15.76085 15.76079
#> 18 15.16307 15.15950 14.96802 14.96805 14.96807 14.96808 14.96808 14.96809
#> 19 17.26205 17.25731 16.54048 16.54055 16.54060 16.54063 16.54065 16.54066
#> 20 14.00696 13.92984 14.02844 14.02838 14.02835 14.02833 14.02831 14.02830
#> 21 25.22808 25.22141 25.30463 25.30488 25.30504 25.30515 25.30522 25.30526
#>      0.1691   0.1577   0.1470   0.1371   0.1279   0.1193   0.1112   0.1037
#> 1  22.28820 22.28820 22.28820 22.28820 23.59097 23.59066 23.59040 23.59018
#> 2  26.04497 26.04497 26.04497 26.04497 25.79695 25.79689 25.79685 25.79680
#> 3  20.94658 20.94656 20.94656 20.94655 20.32587 20.32588 20.32589 20.32590
#> 4  17.06848 17.06847 17.06846 17.06845 17.29648 17.29650 17.29653 17.29655
#> 5  20.63157 20.63155 20.63154 20.63154 20.16385 20.16379 20.16374 20.16370
#> 6  15.17764 15.17762 15.17761 15.17760 14.52429 14.52457 14.52480 14.52500
#> 7  21.01422 21.01421 21.01421 21.01421 21.67924 21.67890 21.67861 21.67837
#> 8  25.71003 25.71005 25.71006 25.71007 25.69558 25.69587 25.69611 25.69630
#> 9  18.43241 18.43241 18.43241 18.43242 18.41048 18.41029 18.41013 18.41000
#> 10 19.27429 19.27430 19.27430 19.27430 19.21673 19.21662 19.21652 19.21644
#> 11 15.63484 15.63483 15.63483 15.63483 16.03478 16.03473 16.03468 16.03465
#> 12 16.88060 16.88059 16.88058 16.88058 17.12881 17.12885 17.12889 17.12892
#> 13 12.40063 12.40066 12.40068 12.40070 12.59617 12.59618 12.59619 12.59619
#> 14 28.10549 28.10550 28.10550 28.10550 28.12318 28.12317 28.12315 28.12314
#> 15 30.32965 30.32967 30.32967 30.32968 30.29106 30.29120 30.29132 30.29142
#> 16 30.05731 30.05731 30.05732 30.05732 29.90175 29.90190 29.90202 29.90212
#> 17 15.76075 15.76072 15.76070 15.76069 16.42067 16.42041 16.42019 16.42000
#> 18 14.96809 14.96809 14.96809 14.96809 14.40729 14.40759 14.40786 14.40808
#> 19 16.54067 16.54068 16.54068 16.54068 16.65478 16.65478 16.65477 16.65477
#> 20 14.02830 14.02829 14.02829 14.02829 13.34454 13.34470 13.34482 13.34492
#> 21 25.30529 25.30531 25.30532 25.30532 24.99653 24.99653 24.99654 24.99654
#>      0.0967   0.0902   0.0841   0.0785   0.0732   0.0682   0.0636   0.0594
#> 1  23.58999 23.58984 23.58971 23.58960 23.41104 23.41097 23.41093 23.77213
#> 2  25.79677 25.79674 25.79671 25.79669 25.73588 25.73592 25.73594 25.46023
#> 3  20.32591 20.32592 20.32592 20.32593 20.68684 20.68665 20.68646 20.85732
#> 4  17.29657 17.29658 17.29660 17.29661 17.69986 17.69962 17.69938 17.72742
#> 5  20.16366 20.16362 20.16360 20.16357 20.25062 20.25054 20.25044 20.31804
#> 6  14.52517 14.52531 14.52543 14.52553 14.48239 14.48259 14.48275 14.59212
#> 7  21.67816 21.67799 21.67785 21.67773 21.73982 21.73956 21.73934 21.65377
#> 8  25.69646 25.69660 25.69671 25.69680 25.73180 25.73181 25.73183 25.73143
#> 9  18.40989 18.40980 18.40973 18.40967 18.18744 18.18748 18.18755 18.22722
#> 10 19.21638 19.21632 19.21628 19.21624 19.02742 19.02747 19.02754 19.13975
#> 11 16.03461 16.03459 16.03456 16.03454 15.60356 15.60379 15.60401 15.32212
#> 12 17.12895 17.12897 17.12899 17.12900 16.79451 16.79472 16.79491 16.66359
#> 13 12.59619 12.59620 12.59620 12.59620 12.90601 12.90586 12.90572 13.08691
#> 14 28.12312 28.12311 28.12310 28.12309 28.05769 28.05773 28.05776 27.91152
#> 15 30.29151 30.29158 30.29165 30.29170 30.41834 30.41838 30.41842 30.69949
#> 16 29.90221 29.90228 29.90234 29.90238 29.90584 29.90590 29.90595 29.89203
#> 17 16.41985 16.41972 16.41961 16.41952 16.56346 16.56322 16.56299 16.52682
#> 18 14.40827 14.40843 14.40857 14.40868 14.22609 14.22642 14.22672 14.05545
#> 19 16.65477 16.65477 16.65476 16.65476 17.25593 17.25556 17.25521 16.87887
#> 20 13.34501 13.34508 13.34514 13.34519 13.05733 13.05757 13.05777 13.37290
#> 21 24.99655 24.99655 24.99655 24.99655 24.85811 24.85824 24.85837 24.71087
#>      0.0554   0.0516   0.0481   0.0449   0.0419   0.0391   0.0364   0.0340
#> 1  23.77217 23.77220 23.77223 23.77226 23.77228 23.69461 23.69465 23.69469
#> 2  25.46011 25.46001 25.45991 25.45981 25.45973 25.24891 25.24888 25.24886
#> 3  20.85756 20.85778 20.85799 20.85819 20.85837 20.91507 20.91528 20.91547
#> 4  17.72754 17.72765 17.72775 17.72785 17.72794 17.74594 17.74604 17.74614
#> 5  20.31816 20.31828 20.31838 20.31849 20.31858 20.48237 20.48244 20.48250
#> 6  14.59227 14.59241 14.59255 14.59267 14.59279 14.30400 14.30424 14.30446
#> 7  21.65358 21.65341 21.65325 21.65310 21.65295 21.14660 21.14657 21.14656
#> 8  25.73144 25.73145 25.73145 25.73146 25.73147 25.65166 25.65168 25.65169
#> 9  18.22708 18.22695 18.22683 18.22672 18.22661 18.40151 18.40133 18.40115
#> 10 19.13968 19.13961 19.13954 19.13948 19.13942 19.40160 19.40144 19.40130
#> 11 15.32186 15.32161 15.32137 15.32115 15.32094 15.51876 15.51845 15.51817
#> 12 16.66346 16.66333 16.66322 16.66311 16.66300 16.86721 16.86703 16.86685
#> 13 13.08706 13.08720 13.08733 13.08746 13.08757 13.05848 13.05862 13.05875
#> 14 27.91144 27.91137 27.91130 27.91124 27.91118 27.97017 27.97008 27.97000
#> 15 30.69964 30.69978 30.69992 30.70004 30.70016 30.81413 30.81423 30.81432
#> 16 29.89207 29.89212 29.89216 29.89219 29.89223 30.03857 30.03857 30.03856
#> 17 16.52679 16.52676 16.52674 16.52672 16.52669 16.53783 16.53780 16.53777
#> 18 14.05537 14.05530 14.05523 14.05516 14.05510 13.87004 13.87003 13.87003
#> 19 16.87880 16.87874 16.87868 16.87862 16.87857 17.07761 17.07750 17.07739
#> 20 13.37312 13.37333 13.37352 13.37371 13.37388 13.32391 13.32413 13.32434
#> 21 24.71080 24.71072 24.71065 24.71059 24.71053 24.53102 24.53101 24.53101
#>      0.0317   0.0295   0.0276   0.0257   0.0240   0.0223   0.0208   0.0194
#> 1  23.69473 23.69476 23.69480 23.69484 23.69488 23.69492 23.69496 23.69500
#> 2  25.24884 25.24883 25.24881 25.24881 25.24880 25.24880 25.24880 25.24880
#> 3  20.91565 20.91581 20.91596 20.91610 20.91623 20.91634 20.91645 20.91655
#> 4  17.74623 17.74631 17.74638 17.74645 17.74652 17.74657 17.74663 17.74668
#> 5  20.48255 20.48260 20.48264 20.48267 20.48270 20.48273 20.48274 20.48276
#> 6  14.30468 14.30488 14.30508 14.30527 14.30545 14.30563 14.30579 14.30595
#> 7  21.14655 21.14656 21.14658 21.14660 21.14663 21.14666 21.14670 21.14674
#> 8  25.65172 25.65174 25.65176 25.65178 25.65180 25.65183 25.65185 25.65188
#> 9  18.40099 18.40083 18.40068 18.40053 18.40040 18.40027 18.40015 18.40003
#> 10 19.40115 19.40102 19.40088 19.40075 19.40063 19.40051 19.40040 19.40029
#> 11 15.51789 15.51764 15.51740 15.51717 15.51695 15.51675 15.51656 15.51638
#> 12 16.86668 16.86652 16.86637 16.86623 16.86609 16.86597 16.86584 16.86573
#> 13 13.05887 13.05899 13.05909 13.05919 13.05929 13.05937 13.05945 13.05953
#> 14 27.96993 27.96985 27.96978 27.96972 27.96966 27.96960 27.96954 27.96949
#> 15 30.81440 30.81448 30.81455 30.81461 30.81467 30.81472 30.81477 30.81481
#> 16 30.03856 30.03855 30.03854 30.03852 30.03851 30.03849 30.03848 30.03846
#> 17 16.53774 16.53772 16.53769 16.53767 16.53765 16.53763 16.53761 16.53759
#> 18 13.87003 13.87003 13.87004 13.87005 13.87006 13.87008 13.87009 13.87011
#> 19 17.07728 17.07717 17.07706 17.07695 17.07685 17.07675 17.07665 17.07655
#> 20 13.32453 13.32472 13.32489 13.32506 13.32521 13.32535 13.32549 13.32562
#> 21 24.53101 24.53101 24.53101 24.53102 24.53102 24.53103 24.53104 24.53105
#>      0.0181   0.0169   0.0158   0.0147   0.0137   0.0128   0.0119   0.0111
#> 1  23.69504 23.69508 23.69512 23.69516 23.69519 23.69523 23.69527 23.69530
#> 2  25.24880 25.24881 25.24882 25.24883 25.24884 25.24885 25.24886 25.24887
#> 3  20.91664 20.91672 20.91679 20.91686 20.91692 20.91697 20.91702 20.91706
#> 4  17.74672 17.74677 17.74680 17.74684 17.74687 17.74690 17.74692 17.74695
#> 5  20.48277 20.48278 20.48278 20.48278 20.48278 20.48278 20.48277 20.48277
#> 6  14.30611 14.30625 14.30639 14.30652 14.30664 14.30676 14.30687 14.30698
#> 7  21.14679 21.14684 21.14689 21.14695 21.14700 21.14706 21.14712 21.14718
#> 8  25.65190 25.65192 25.65195 25.65197 25.65200 25.65202 25.65204 25.65206
#> 9  18.39992 18.39982 18.39972 18.39963 18.39954 18.39946 18.39939 18.39931
#> 10 19.40019 19.40009 19.40000 19.39991 19.39982 19.39974 19.39966 19.39959
#> 11 15.51621 15.51605 15.51591 15.51577 15.51564 15.51552 15.51540 15.51530
#> 12 16.86562 16.86552 16.86542 16.86533 16.86525 16.86517 16.86509 16.86502
#> 13 13.05960 13.05966 13.05972 13.05977 13.05982 13.05987 13.05991 13.05995
#> 14 27.96944 27.96939 27.96934 27.96930 27.96926 27.96922 27.96919 27.96915
#> 15 30.81485 30.81488 30.81491 30.81494 30.81496 30.81498 30.81500 30.81502
#> 16 30.03844 30.03842 30.03840 30.03838 30.03836 30.03834 30.03832 30.03831
#> 17 16.53757 16.53756 16.53754 16.53753 16.53751 16.53750 16.53749 16.53748
#> 18 13.87013 13.87015 13.87017 13.87020 13.87022 13.87024 13.87027 13.87029
#> 19 17.07646 17.07636 17.07627 17.07619 17.07610 17.07602 17.07594 17.07587
#> 20 13.32574 13.32585 13.32595 13.32605 13.32614 13.32623 13.32631 13.32638
#> 21 24.53107 24.53108 24.53109 24.53111 24.53112 24.53114 24.53116 24.53117
#>      0.0104   0.0097   0.0090   0.0084   0.0078   0.0073   0.0068   0.0064
#> 1  23.69533 23.69537 23.69540 23.69543 23.69546 23.69549 23.69551 23.69554
#> 2  25.24888 25.24890 25.24891 25.24892 25.24894 25.24895 25.24897 25.24898
#> 3  20.91710 20.91713 20.91716 20.91718 20.91721 20.91722 20.91724 20.91725
#> 4  17.74697 17.74699 17.74700 17.74702 17.74703 17.74704 17.74705 17.74706
#> 5  20.48276 20.48275 20.48274 20.48272 20.48271 20.48270 20.48268 20.48267
#> 6  14.30708 14.30718 14.30727 14.30735 14.30743 14.30751 14.30758 14.30765
#> 7  21.14724 21.14730 21.14736 21.14742 21.14748 21.14754 21.14759 21.14765
#> 8  25.65208 25.65211 25.65213 25.65215 25.65217 25.65218 25.65220 25.65222
#> 9  18.39925 18.39918 18.39912 18.39907 18.39901 18.39896 18.39892 18.39888
#> 10 19.39952 19.39945 19.39939 19.39933 19.39927 19.39922 19.39917 19.39912
#> 11 15.51520 15.51511 15.51502 15.51495 15.51487 15.51480 15.51474 15.51468
#> 12 16.86495 16.86489 16.86483 16.86478 16.86473 16.86468 16.86464 16.86459
#> 13 13.05999 13.06002 13.06005 13.06007 13.06010 13.06012 13.06014 13.06016
#> 14 27.96912 27.96909 27.96906 27.96904 27.96901 27.96899 27.96897 27.96895
#> 15 30.81503 30.81504 30.81505 30.81506 30.81507 30.81507 30.81508 30.81508
#> 16 30.03829 30.03827 30.03825 30.03823 30.03821 30.03819 30.03817 30.03816
#> 17 16.53747 16.53746 16.53745 16.53745 16.53744 16.53743 16.53743 16.53742
#> 18 13.87031 13.87034 13.87036 13.87038 13.87041 13.87043 13.87045 13.87048
#> 19 17.07579 17.07572 17.07566 17.07559 17.07553 17.07547 17.07541 17.07536
#> 20 13.32645 13.32651 13.32657 13.32663 13.32668 13.32672 13.32677 13.32681
#> 21 24.53119 24.53121 24.53122 24.53124 24.53126 24.53127 24.53129 24.53130
#>      0.0059   0.0055   0.0052   0.0048
#> 1  23.69557 23.69559 23.69561 23.69564
#> 2  25.24900 25.24901 25.24902 25.24904
#> 3  20.91726 20.91727 20.91728 20.91728
#> 4  17.74707 17.74707 17.74708 17.74708
#> 5  20.48265 20.48264 20.48262 20.48261
#> 6  14.30771 14.30777 14.30783 14.30788
#> 7  21.14771 21.14776 21.14781 21.14787
#> 8  25.65224 25.65225 25.65227 25.65229
#> 9  18.39884 18.39880 18.39876 18.39873
#> 10 19.39907 19.39903 19.39899 19.39895
#> 11 15.51463 15.51458 15.51454 15.51449
#> 12 16.86456 16.86452 16.86449 16.86446
#> 13 13.06017 13.06019 13.06020 13.06021
#> 14 27.96893 27.96891 27.96889 27.96888
#> 15 30.81508 30.81508 30.81508 30.81508
#> 16 30.03814 30.03812 30.03811 30.03809
#> 17 16.53741 16.53741 16.53740 16.53740
#> 18 13.87050 13.87052 13.87054 13.87056
#> 19 17.07531 17.07526 17.07521 17.07517
#> 20 13.32684 13.32688 13.32691 13.32694
#> 21 24.53132 24.53133 24.53135 24.53136
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 22.8 21.4 18.7 18.1 14.3 24.4 22.8 19.2 17.8 16.4 17.3 14.7 32.4 30.4
#> [16] 33.9 15.5 13.3 15.8 15.0 21.4
#> 
#> $X
#>               am       carb        cyl        disp        drat       gear
#>  [1,]  1.2747549  0.7210872 -0.1097643 -0.60653852  0.48458390  0.5188745
#>  [2,]  1.2747549 -1.0960525 -1.2622890 -1.08242591  0.39274038  0.5188745
#>  [3,] -0.7844645 -1.0960525 -0.1097643  0.29032617 -1.02164981 -1.0377490
#>  [4,] -0.7844645 -0.4903393  1.0427605  1.22379759 -0.89306888 -1.0377490
#>  [5,] -0.7844645 -1.0960525 -0.1097643 -0.01167929 -1.60944833 -1.0377490
#>  [6,] -0.7844645  0.7210872  1.0427605  1.22379759 -0.78285666 -1.0377490
#>  [7,] -0.7844645 -0.4903393 -1.2622890 -0.72825587  0.09884112  0.5188745
#>  [8,] -0.7844645 -0.4903393 -1.2622890 -0.78225079  0.52132131  0.5188745
#>  [9,] -0.7844645  0.7210872 -0.1097643 -0.53698575  0.52132131  0.5188745
#> [10,] -0.7844645  0.7210872 -0.1097643 -0.53698575  0.52132131  0.5188745
#> [11,] -0.7844645  0.1153739  1.0427605  0.45322608 -1.04001851 -1.0377490
#> [12,] -0.7844645  0.1153739  1.0427605  0.45322608 -1.04001851 -1.0377490
#> [13,] -0.7844645  0.7210872  1.0427605  1.95593203 -0.74611925 -1.0377490
#> [14,]  1.2747549 -1.0960525 -1.2622890 -1.35057015  0.81522057  0.5188745
#> [15,]  1.2747549 -0.4903393 -1.2622890 -1.37802519  2.37656039  0.5188745
#> [16,]  1.2747549 -1.0960525 -1.2622890 -1.42012292  1.07238242  0.5188745
#> [17,] -0.7844645 -0.4903393  1.0427605  0.83942700 -1.60944833 -1.0377490
#> [18,] -0.7844645  0.7210872  1.0427605  1.13228078  0.17231594 -1.0377490
#> [19,]  1.2747549  0.7210872  1.0427605  1.14143246  1.07238242  2.0754981
#> [20,]  1.2747549  3.1439401  1.0427605  0.68384843 -0.17668944  2.0754981
#> [21,]  1.2747549 -0.4903393 -1.2622890 -0.96345406  0.87032668  0.5188745
#>                hp       qsec         vs            wt
#>  [1,] -0.50770444 -0.7725909 -1.0488088 -0.7472087103
#>  [2,] -0.73230231  0.3066937  0.9534626 -1.1238019003
#>  [3,] -0.50770444  0.7233478  0.9534626 -0.0002988835
#>  [4,]  0.35105214 -0.4914749 -1.0488088  0.2821460090
#>  [5,] -0.57376264  1.1149022  0.9534626  0.3072522217
#>  [6,]  1.27586692 -1.0838265 -1.0488088  0.4453363913
#>  [7,] -1.14186315  1.0044638  0.9534626 -0.0316816493
#>  [8,] -0.70587904  2.4602431  0.9534626 -0.0818940746
#>  [9,] -0.33595312  0.1510759  0.9534626  0.2821460090
#> [10,] -0.33595312  0.4522717  0.9534626  0.2821460090
#> [11,]  0.41711034 -0.3007176 -1.0488088  1.0729917080
#> [12,]  0.41711034 -0.2003191 -1.0488088  0.6461860927
#> [13,]  1.07769232 -0.2906778 -1.0488088  2.6735127654
#> [14,] -1.08901659  0.7384076  0.9534626 -1.2744391763
#> [15,] -1.27397954  0.2615144  0.9534626 -2.0087958967
#> [16,] -1.10222823  0.9542645  0.9534626 -1.7326275574
#> [17,]  0.02076115 -0.5667738 -1.0488088  0.3825708597
#> [18,]  1.27586692 -1.2996834 -1.0488088  0.7842702623
#> [19,]  1.52688807 -1.7564969 -1.0488088 -0.0567878620
#> [20,]  2.46491449 -1.7062976 -1.0488088  0.4453363913
#> [21,] -0.52091608  0.3016738  0.9534626 -0.5463590090
#> attr(,"center")
#>  [1]   0.3809524   2.8095238   6.1904762 226.2761905   3.6361905   3.6666667
#>  [7] 148.4285714  17.9990476   0.5238095   3.2152381
#> attr(,"scale")
#>  [1]   0.4856209   1.6509463   1.7353207 109.2695483   0.5444042   0.6424161
#>  [7]  75.6908319   1.9920602   0.4994328   0.7966156
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 30
#> 
#> $lambda.min
#> [1] 0.6364791
#> 
#> $null.dev
#> [1] 32.9717
#> 
#> $Bias
#> [1] 8.358221
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 4.685582