Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 35.453668 33.240023 29.875780 26.630258 23.945356 21.738760 19.939504
#>   [8] 18.486449 17.326963 16.415778 15.713998 15.188233 14.809853 14.537694
#>  [15] 14.348975 14.137839 13.507138 12.939036 12.635136 12.407903 12.210735
#>  [22] 12.039779 11.829077 11.522821 11.012749 10.956080 10.827232 10.776633
#>  [29]  9.112986  9.084357  9.079168  9.081249  9.089727  9.035514  8.882670
#>  [36]  8.758063  8.703824  9.216525  9.469688  9.783018 10.116498 10.725222
#>  [43] 10.887365 10.870727 10.673911 10.952354 10.955909 10.958079 10.510831
#>  [50] 11.155284 11.209937 11.227031 19.609910 19.798043 20.788516 20.790574
#>  [57] 20.810365 20.898099 20.885750 20.887477 20.889340 20.997475 20.865147
#>  [64] 21.327794 23.045077 23.305435 22.924402 22.912382 22.912548 23.028696
#>  [71] 22.926273 22.865298 22.865413 22.859514 22.859554 22.859718 23.266753
#>  [78] 23.266874 23.378074 23.378478 23.474477 23.474822 23.475115 23.476119
#>  [85] 23.508527 23.508698 23.508665 23.508350 23.413732 23.611569 23.611161
#>  [92] 23.610724 23.610261 23.610000 23.620189 23.619866 23.627351 23.627053
#>  [99] 23.652931 23.652993
#> 
#> $cvse
#>   [1] 11.122292 11.079554 10.456641  9.560082  8.754550  8.029343  7.375734
#>   [8]  6.786637  6.256323  5.780172  5.354456  4.976139  4.642698  4.352333
#>  [15]  4.101348  3.885938  3.680096  3.502873  3.458591  3.459644  3.464021
#>  [22]  3.462381  3.455951  3.439239  3.215239  3.028529  2.918132  2.860918
#>  [29]  2.694532  2.707999  2.718305  2.726385  2.725222  2.703645  2.676220
#>  [36]  2.653228  2.635637  2.596752  2.710443  2.846829  3.010018  3.191158
#>  [43]  3.271512  3.264574  3.262331  3.221586  3.221349  3.221361  2.986338
#>  [50]  3.269504  3.272663  3.273842  9.219719  9.255691  9.330638  9.331683
#>  [57]  9.331056  9.328163  9.328664  9.328566  9.329434  9.327642  9.342224
#>  [64]  9.366683  9.503437  9.489370  9.145262  9.145984  9.145379  9.139973
#>  [71]  9.124245  9.127983  9.127421  9.127601  9.127075  9.126561  9.118808
#>  [78]  9.118333  9.058628  9.058197  9.073353  9.072957  9.072574  9.072159
#>  [85]  9.070007  9.069661  9.069719  9.069388  9.076661  9.100751  9.100447
#>  [92]  9.100150  9.099861  9.099566  9.098702  9.098422  9.097571  9.097290
#>  [99]  9.090681  9.090395
#> 
#> $fold
#>  [1]  5  2 10  1  7  2  3  1  6  7  9  8  6  3  4  1  4 10  5  8  9
#> 
#> $lambda
#>   [1] 4.995091388 4.658438946 4.344475753 4.051672627 3.778603453 3.523938227
#>   [7] 3.286436584 3.064941758 2.858374941 2.665730037 2.486068754 2.318516041
#>  [13] 2.162255820 2.016527014 1.880619842 1.753872359 1.635667232 1.525428735
#>  [19] 1.422619944 1.326740121 1.237322277 1.153930897 1.076159817 1.003630247
#>  [25] 0.935988927 0.872906406 0.814075436 0.759209476 0.708041298 0.660321685
#>  [31] 0.615818213 0.574314127 0.535607277 0.499509139 0.465843895 0.434447575
#>  [37] 0.405167263 0.377860345 0.352393823 0.328643658 0.306494176 0.285837494
#>  [43] 0.266573004 0.248606875 0.231851604 0.216225582 0.201652701 0.188061984
#>  [49] 0.175387236 0.163566723 0.152542874 0.142261994 0.132674012 0.123732228
#>  [55] 0.115393090 0.107615982 0.100363025 0.093598893 0.087290641 0.081407544
#>  [61] 0.075920948 0.070804130 0.066032168 0.061581821 0.057431413 0.053560728
#>  [67] 0.049950914 0.046584389 0.043444758 0.040516726 0.037786035 0.035239382
#>  [73] 0.032864366 0.030649418 0.028583749 0.026657300 0.024860688 0.023185160
#>  [79] 0.021622558 0.020165270 0.018806198 0.017538724 0.016356672 0.015254287
#>  [85] 0.014226199 0.013267401 0.012373223 0.011539309 0.010761598 0.010036302
#>  [91] 0.009359889 0.008729064 0.008140754 0.007592095 0.007080413 0.006603217
#>  [97] 0.006158182 0.005743141 0.005356073 0.004995091
#> 
#> $fit
#> $beta
#>               4.9951     4.6584     4.3445     4.0517    3.7786    3.5239
#> (Intercept) 20.10952 21.8877768 23.5461815 25.0928152 26.535211 27.880394
#> am           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> cyl          0.00000 -0.2786814 -0.5385807 -0.7809636 -1.007011 -1.217823
#> disp         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> wt           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#>                3.2864    3.0649    2.8584    2.6657    2.4861    2.3185
#> (Intercept) 29.134917 30.304888 31.396008 32.413590 33.362590 34.247630
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.414427 -1.597781 -1.768777 -1.928249 -2.076973 -2.215673
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                2.1623   2.0165    1.8806    1.7539   1.6357   1.5254   1.4226
#> (Intercept) 35.073022 35.84279 36.560669 37.230170 37.69945 37.69945 37.69945
#> am           0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.00000
#> carb         0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.00000
#> cyl         -2.345026 -2.46566 -2.578165 -2.683086 -2.75663 -2.75663 -2.75663
#> disp         0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.00000
#> drat         0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.00000
#> gear         0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.00000
#> hp           0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.00000
#> qsec         0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.00000
#> vs           0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.00000
#> wt           0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.00000
#>               1.3267   1.2373   1.1539     1.0762    1.0036    0.9360    0.8729
#> (Intercept) 37.69945 37.69945 37.69945 37.8628075 38.945408 39.954908 40.178920
#> am           0.00000  0.00000  0.00000  0.0000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.00000  0.00000  0.0000000  0.000000  0.000000  0.000000
#> cyl         -2.75663 -2.75663 -2.75663 -2.6793231 -2.160526 -1.676797 -1.569185
#> disp         0.00000  0.00000  0.00000  0.0000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.00000  0.00000  0.0000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.00000  0.00000  0.0000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.00000  0.00000  0.0000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.00000  0.00000  0.0000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.00000  0.00000  0.0000000  0.000000  0.000000  0.000000
#> wt           0.00000  0.00000  0.00000 -0.2089216 -1.606617 -2.909859 -3.199603
#>                0.8141    0.7592      0.7080    0.6603     0.6158    0.5743
#> (Intercept) 40.178699 40.178565 40.18595386 40.206448 40.2255417 40.243359
#> am           0.000000  0.000000  0.00000000  0.000000  0.0000000  0.000000
#> carb         0.000000  0.000000 -0.02243094 -0.083227 -0.1398193 -0.192704
#> cyl         -1.569007 -1.568899 -1.55941245 -1.534076 -1.5105050 -1.488457
#> disp         0.000000  0.000000  0.00000000  0.000000  0.0000000  0.000000
#> drat         0.000000  0.000000  0.00000000  0.000000  0.0000000  0.000000
#> gear         0.000000  0.000000  0.00000000  0.000000  0.0000000  0.000000
#> hp           0.000000  0.000000  0.00000000  0.000000  0.0000000  0.000000
#> qsec         0.000000  0.000000  0.00000000  0.000000  0.0000000  0.000000
#> vs           0.000000  0.000000  0.00000000  0.000000  0.0000000  0.000000
#> wt          -3.199893 -3.200069 -3.20060973 -3.201459 -3.2022288 -3.202982
#>                 0.5356     0.4995     0.4658     0.4344     0.4052     0.3779
#> (Intercept) 40.2599731 40.2754678 40.2899181 40.3033945 40.3159627 40.3276838
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb        -0.2420262 -0.2880245 -0.3309226 -0.3709296 -0.4082402 -0.4430362
#> cyl         -1.4678931 -1.4487146 -1.4308287 -1.4141482 -1.3985919 -1.3840841
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -3.2036866 -3.2043443 -3.2049577 -3.2055298 -3.2060633 -3.2065609
#>                 0.3524     0.3286    0.3065     0.2858     0.2666     0.2486
#> (Intercept) 40.3386150 40.3488094 40.358317 40.3603784 24.3835342 24.3811434
#> am           0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> carb        -0.4754871 -0.5057509 -0.533975 -0.5401278  0.0000000  0.0000000
#> cyl         -1.3705541 -1.3579359 -1.346168 -1.3435940 -0.6408241 -0.6405704
#> disp         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0000000  0.0000000  0.000000  0.0000000  0.7749445  0.7750543
#> vs           0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> wt          -3.2070249 -3.2074577 -3.207861 -3.2079635 -4.3854703 -4.3858380
#>                 0.2319     0.2162     0.2017     0.1881     0.1754      0.1636
#> (Intercept) 24.3792053 24.3776342 24.3763606 24.3753281 25.0792971 25.91352421
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.00000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000 -0.0441003 -0.09468469
#> cyl         -0.6403647 -0.6401980 -0.6400628 -0.6399533 -0.6610214 -0.68799992
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.00000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.00000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.00000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.00000000
#> qsec         0.7751434  0.7752156  0.7752741  0.7753215  0.7414436  0.70135383
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.00000000
#> wt          -4.3861360 -4.3863777 -4.3865735 -4.3867323 -4.3373665 -4.27667212
#>                 0.1525     0.1423     0.1327     0.1237     0.1154     0.1076
#> (Intercept) 19.3465018 19.3446557 19.3431953 19.3420450 19.3411433 24.8338839
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb        -0.3088520 -0.3087752 -0.3087099 -0.3086544 -0.3086075  0.0000000
#> cyl         -0.5036148 -0.5033919 -0.5032079 -0.5030563 -0.5029318 -0.7949506
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         1.3853775  1.3851642  1.3849724  1.3848016  1.3846505  1.8597947
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 -1.0949985
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.6590186  0.6591578  0.6592736  0.6593698  0.6594495  0.5457493
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -3.7163353 -3.7168047 -3.7172015 -3.7175360 -3.7178171 -3.7631268
#>                 0.1004     0.0936     0.0873     0.0814     0.0759       0.0708
#> (Intercept) 24.8285804 24.8238005 24.8195115 24.8156788 24.8122670 21.363449340
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000000
#> cyl         -0.7948040 -0.7946529 -0.7945015 -0.7943531 -0.7942100 -0.470641565
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000000
#> drat         1.8605655  1.8611709  1.8616400  1.8619977  1.8622647  1.888033351
#> gear        -1.0950356 -1.0950320 -1.0949980 -1.0949420 -1.0948708 -0.626757310
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 -0.008515165
#> qsec         0.5458151  0.5458879  0.5459642  0.5460417  0.5461184  0.561302044
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000000
#> wt          -3.7629458 -3.7628379 -3.7627870 -3.7627795 -3.7628046 -3.588391954
#>                   0.0660      0.0616       0.0574       0.0536       0.0500
#> (Intercept) 21.360756305 16.92909827 16.933901610 16.938237428 16.942130942
#> am           0.000000000  0.94728743  0.946590759  0.945910518  0.945249366
#> carb         0.000000000  0.00000000  0.000000000  0.000000000  0.000000000
#> cyl         -0.470374333 -0.38170434 -0.381965943 -0.382195017 -0.382393809
#> disp         0.000000000  0.00948494  0.009482766  0.009480521  0.009478227
#> drat         1.887432003  1.02174748  1.021791641  1.021875239  1.021993126
#> gear        -0.626389526  0.00000000  0.000000000  0.000000000  0.000000000
#> hp          -0.008514284 -0.01472981 -0.014727975 -0.014726409 -0.014725092
#> qsec         0.561493743  0.83064195  0.830444894  0.830257052  0.830078617
#> vs           0.000000000  0.00000000  0.000000000  0.000000000  0.000000000
#> wt          -3.588939307 -4.16404359 -4.163812874 -4.163580446 -4.163348065
#>                  0.0466      0.0434      0.0405      0.0378      0.0352
#> (Intercept) 19.20972545 19.21476097 19.21953903 19.22406686 19.25052343
#> am           1.01355074  1.01322176  1.01288908  1.01255351  0.88270236
#> carb         0.24185726  0.24176145  0.24166419  0.24156575  0.25808908
#> cyl         -0.56567933 -0.56596364 -0.56622977 -0.56647836 -0.39592973
#> disp         0.01263959  0.01263533  0.01263103  0.01262668  0.01202250
#> drat         1.04143174  1.04179057  1.04215879  1.04253520  1.07011162
#> gear        -0.47326538 -0.47368910 -0.47409297 -0.47447748 -0.32953793
#> hp          -0.01601669 -0.01601062 -0.01600476 -0.01599910 -0.01901757
#> qsec         0.85259590  0.85237646  0.85216113  0.85195005  0.74918356
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.69805800
#> wt          -4.50397617 -4.50351770 -4.50305605 -4.50259224 -4.35834866
#>                  0.0329      0.0306      0.0286      0.0267      0.0249
#> (Intercept) 19.25047653 19.25049667 19.25058064 19.25072531 19.25092764
#> am           0.88304548  0.88338155  0.88371054  0.88403241  0.88434713
#> carb         0.25819761  0.25830262  0.25840418  0.25850232  0.25859709
#> cyl         -0.39601223 -0.39610034 -0.39619368 -0.39629189 -0.39639464
#> disp         0.01202719  0.01203176  0.01203619  0.01204051  0.01204470
#> drat         1.06964571  1.06919178  1.06874977  1.06831962  1.06790125
#> gear        -0.32951785 -0.32950670 -0.32950408 -0.32950959 -0.32952283
#> hp          -0.01901913 -0.01902052 -0.01902175 -0.01902281 -0.01902373
#> qsec         0.74932015  0.74945340  0.74958332  0.74970991  0.74983320
#> vs           0.69792247  0.69777759  0.69762396  0.69746218  0.69729280
#> wt          -4.35883140 -4.35930214 -4.35976091 -4.36020777 -4.36064278
#>                  0.0232      0.0216      0.0202      0.0188      0.0175
#> (Intercept) 19.25118466 19.25149348 19.25185130 19.25225539 19.25270309
#> am           0.88465469  0.88495508  0.88524830  0.88553436  0.88581329
#> carb         0.25868856  0.25877677  0.25886178  0.25894365  0.25902242
#> cyl         -0.39650160 -0.39661244 -0.39672685 -0.39684454 -0.39696522
#> disp         0.01204876  0.01205271  0.01205654  0.01206024  0.01206383
#> drat         1.06749457  1.06709949  1.06671590  1.06634369  1.06598274
#> gear        -0.32954341 -0.32957098 -0.32960517 -0.32964561 -0.32969198
#> hp          -0.01902450 -0.01902513 -0.01902563 -0.01902600 -0.01902625
#> qsec         0.74995319  0.75006992  0.75018341  0.75029367  0.75040075
#> vs           0.69711639  0.69693347  0.69674456  0.69655016  0.69635075
#> wt          -4.36106601 -4.36147755 -4.36187748 -4.36226590 -4.36264293
#>                  0.0164      0.0153      0.0142      0.0133      0.0124
#> (Intercept) 19.25319182 19.25371909 19.25428246 19.25487960 19.25550822
#> am           0.88608509  0.88634981  0.88660748  0.88685815  0.88710186
#> carb         0.25909817  0.25917095  0.25924081  0.25930782  0.25937204
#> cyl         -0.39708860 -0.39721443 -0.39734244 -0.39747239 -0.39760403
#> disp         0.01206731  0.01207067  0.01207392  0.01207706  0.01208008
#> drat         1.06563292  1.06529410  1.06496613  1.06464887  1.06434218
#> gear        -0.32974394 -0.32980115 -0.32986332 -0.32993012 -0.33000128
#> hp          -0.01902638 -0.01902640 -0.01902631 -0.01902613 -0.01902585
#> qsec         0.75050466  0.75060545  0.75070315  0.75079779  0.75088941
#> vs           0.69614680  0.69593874  0.69572701  0.69551202  0.69529417
#> wt          -4.36300868 -4.36336326 -4.36370682 -4.36403948 -4.36436139
#>                  0.0115      0.0108      0.0100      0.0094      0.0087
#> (Intercept) 19.25616612 19.25685116 19.25756128 19.25829450 19.25904888
#> am           0.88733868  0.88756866  0.88779186  0.88800836  0.88821823
#> carb         0.25943353  0.25949234  0.25954854  0.25960219  0.25965335
#> cyl         -0.39773714 -0.39787150 -0.39800689 -0.39814312 -0.39827999
#> disp         0.01208301  0.01208582  0.01208853  0.01209114  0.01209365
#> drat         1.06404589  1.06375985  1.06348389  1.06321784  1.06296154
#> gear        -0.33007650 -0.33015549 -0.33023801 -0.33032377 -0.33041254
#> hp          -0.01902547 -0.01902501 -0.01902447 -0.01902385 -0.01902316
#> qsec         0.75097806  0.75106377  0.75114659  0.75122655  0.75130372
#> vs           0.69507384  0.69485139  0.69462719  0.69440155  0.69417482
#> wt          -4.36467270 -4.36497356 -4.36526413 -4.36554455 -4.36581502
#>                  0.0081      0.0076      0.0071      0.0066      0.0062
#> (Intercept) 19.25982257 19.26061378 19.26142078 19.26224192 19.26307559
#> am           0.88842155  0.88861839  0.88880886  0.88899302  0.88917099
#> carb         0.25970207  0.25974843  0.25979248  0.25983428  0.25987389
#> cyl         -0.39841732 -0.39855492 -0.39869264 -0.39883031 -0.39896777
#> disp         0.01209607  0.01209838  0.01210061  0.01210274  0.01210478
#> drat         1.06271481  1.06247747  1.06224935  1.06203026  1.06182003
#> gear        -0.33050407 -0.33059812 -0.33069448 -0.33079291 -0.33089322
#> hp          -0.01902240 -0.01902158 -0.01902070 -0.01901975 -0.01901876
#> qsec         0.75137812  0.75144982  0.75151885  0.75158526  0.75164912
#> vs           0.69394729  0.69371926  0.69349101  0.69326281  0.69303492
#> wt          -4.36607568 -4.36632671 -4.36656828 -4.36680058 -4.36702378
#>                  0.0057      0.0054      0.0050
#> (Intercept) 19.26392026 19.26477445 19.26563676
#> am           0.88934284  0.88950867  0.88966859
#> carb         0.25991137  0.25994677  0.25998017
#> cyl         -0.39910489 -0.39924152 -0.39937753
#> disp         0.01210673  0.01210859  0.01211038
#> drat         1.06161847  1.06142540  1.06124062
#> gear        -0.33099519 -0.33109864 -0.33120338
#> hp          -0.01901772 -0.01901663 -0.01901550
#> qsec         0.75171045  0.75176932  0.75182577
#> vs           0.69280757  0.69258100  0.69235542
#> wt          -4.36723805 -4.36744359 -4.36764058
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1  37  57  56  12   1   1   6   6   5   5   5   5   5   5
#>  [37]   5   5   5   5   5   3  53   1   1   1   1   1  28  30  70   1   1   1
#>  [55]   1  42   1   1   1   1   1  64   1 157   1   1   1 103   1   1   1 122
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.995091388 4.658438946 4.344475753 4.051672627 3.778603453 3.523938227
#>   [7] 3.286436584 3.064941758 2.858374941 2.665730037 2.486068754 2.318516041
#>  [13] 2.162255820 2.016527014 1.880619842 1.753872359 1.635667232 1.525428735
#>  [19] 1.422619944 1.326740121 1.237322277 1.153930897 1.076159817 1.003630247
#>  [25] 0.935988927 0.872906406 0.814075436 0.759209476 0.708041298 0.660321685
#>  [31] 0.615818213 0.574314127 0.535607277 0.499509139 0.465843895 0.434447575
#>  [37] 0.405167263 0.377860345 0.352393823 0.328643658 0.306494176 0.285837494
#>  [43] 0.266573004 0.248606875 0.231851604 0.216225582 0.201652701 0.188061984
#>  [49] 0.175387236 0.163566723 0.152542874 0.142261994 0.132674012 0.123732228
#>  [55] 0.115393090 0.107615982 0.100363025 0.093598893 0.087290641 0.081407544
#>  [61] 0.075920948 0.070804130 0.066032168 0.061581821 0.057431413 0.053560728
#>  [67] 0.049950914 0.046584389 0.043444758 0.040516726 0.037786035 0.035239382
#>  [73] 0.032864366 0.030649418 0.028583749 0.026657300 0.024860688 0.023185160
#>  [79] 0.021622558 0.020165270 0.018806198 0.017538724 0.016356672 0.015254287
#>  [85] 0.014226199 0.013267401 0.012373223 0.011539309 0.010761598 0.010036302
#>  [91] 0.009359889 0.008729064 0.008140754 0.007592095 0.007080413 0.006603217
#>  [97] 0.006158182 0.005743141 0.005356073 0.004995091
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 22
#> 
#> $loss
#>   [1] 683.73810 583.15176 498.99633 428.90727 370.84316 323.04257 283.98665
#>   [8] 252.36666 227.05575 207.08453 191.61981 179.94620 171.45016 165.60613
#>  [15] 161.96452 160.14134 159.76840 159.76840 159.76840 159.76840 159.76840
#>  [22] 159.76840 151.79778 112.31024  97.17992  96.66001  96.66000  96.66000
#>  [29]  95.53827  92.73715  90.44374  88.57425  87.06899  85.87243  84.93677
#>  [36]  84.22095  83.68974  83.31294  83.06468  82.92287  82.86864  82.86682
#>  [43]  78.07307  78.07306  78.07306  78.07305  78.07305  78.07305  77.54941
#>  [50]  77.08228  73.73097  73.73096  73.73096  73.73095  73.73095  72.43107
#>  [57]  72.43106  72.43106  72.43105  72.43105  72.43105  71.45965  71.45963
#>  [64]  69.99087  69.99086  69.99085  69.99084  69.28548  69.28545  69.28543
#>  [71]  69.28540  69.03941  69.03940  69.03938  69.03937  69.03936  69.03934
#>  [78]  69.03933  69.03932  69.03931  69.03930  69.03929  69.03928  69.03927
#>  [85]  69.03926  69.03926  69.03925  69.03924  69.03923  69.03923  69.03922
#>  [92]  69.03922  69.03921  69.03921  69.03920  69.03920  69.03919  69.03919
#>  [99]  69.03918  69.03918
#> 
#> $linear.predictors
#>      4.9951   4.6584   4.3445   4.0517   3.7786   3.5239   3.2864   3.0649
#> 1  20.10952 20.21569 20.31470 20.40703 20.49315 20.57346 20.64835 20.71820
#> 2  20.10952 20.21569 20.31470 20.40703 20.49315 20.57346 20.64835 20.71820
#> 3  20.10952 20.77305 21.39186 21.96896 22.50717 23.00910 23.47721 23.91376
#> 4  20.10952 19.65833 19.23754 18.84511 18.47913 18.13781 17.81950 17.52264
#> 5  20.10952 19.65833 19.23754 18.84511 18.47913 18.13781 17.81950 17.52264
#> 6  20.10952 20.77305 21.39186 21.96896 22.50717 23.00910 23.47721 23.91376
#> 7  20.10952 20.77305 21.39186 21.96896 22.50717 23.00910 23.47721 23.91376
#> 8  20.10952 19.65833 19.23754 18.84511 18.47913 18.13781 17.81950 17.52264
#> 9  20.10952 19.65833 19.23754 18.84511 18.47913 18.13781 17.81950 17.52264
#> 10 20.10952 19.65833 19.23754 18.84511 18.47913 18.13781 17.81950 17.52264
#> 11 20.10952 19.65833 19.23754 18.84511 18.47913 18.13781 17.81950 17.52264
#> 12 20.10952 20.77305 21.39186 21.96896 22.50717 23.00910 23.47721 23.91376
#> 13 20.10952 20.77305 21.39186 21.96896 22.50717 23.00910 23.47721 23.91376
#> 14 20.10952 19.65833 19.23754 18.84511 18.47913 18.13781 17.81950 17.52264
#> 15 20.10952 19.65833 19.23754 18.84511 18.47913 18.13781 17.81950 17.52264
#> 16 20.10952 19.65833 19.23754 18.84511 18.47913 18.13781 17.81950 17.52264
#> 17 20.10952 20.77305 21.39186 21.96896 22.50717 23.00910 23.47721 23.91376
#> 18 20.10952 20.77305 21.39186 21.96896 22.50717 23.00910 23.47721 23.91376
#> 19 20.10952 19.65833 19.23754 18.84511 18.47913 18.13781 17.81950 17.52264
#> 20 20.10952 20.21569 20.31470 20.40703 20.49315 20.57346 20.64835 20.71820
#> 21 20.10952 19.65833 19.23754 18.84511 18.47913 18.13781 17.81950 17.52264
#>      2.8584   2.6657   2.4861   2.3185   2.1623   2.0165   1.8806   1.7539
#> 1  20.78334 20.84409 20.90075 20.95359 21.00287 21.04882 21.09168 21.13165
#> 2  20.78334 20.84409 20.90075 20.95359 21.00287 21.04882 21.09168 21.13165
#> 3  24.32090 24.70059 25.05470 25.38494 25.69292 25.98014 26.24801 26.49782
#> 4  17.24579 16.98760 16.74681 16.52224 16.31282 16.11750 15.93535 15.76548
#> 5  17.24579 16.98760 16.74681 16.52224 16.31282 16.11750 15.93535 15.76548
#> 6  24.32090 24.70059 25.05470 25.38494 25.69292 25.98014 26.24801 26.49782
#> 7  24.32090 24.70059 25.05470 25.38494 25.69292 25.98014 26.24801 26.49782
#> 8  17.24579 16.98760 16.74681 16.52224 16.31282 16.11750 15.93535 15.76548
#> 9  17.24579 16.98760 16.74681 16.52224 16.31282 16.11750 15.93535 15.76548
#> 10 17.24579 16.98760 16.74681 16.52224 16.31282 16.11750 15.93535 15.76548
#> 11 17.24579 16.98760 16.74681 16.52224 16.31282 16.11750 15.93535 15.76548
#> 12 24.32090 24.70059 25.05470 25.38494 25.69292 25.98014 26.24801 26.49782
#> 13 24.32090 24.70059 25.05470 25.38494 25.69292 25.98014 26.24801 26.49782
#> 14 17.24579 16.98760 16.74681 16.52224 16.31282 16.11750 15.93535 15.76548
#> 15 17.24579 16.98760 16.74681 16.52224 16.31282 16.11750 15.93535 15.76548
#> 16 17.24579 16.98760 16.74681 16.52224 16.31282 16.11750 15.93535 15.76548
#> 17 24.32090 24.70059 25.05470 25.38494 25.69292 25.98014 26.24801 26.49782
#> 18 24.32090 24.70059 25.05470 25.38494 25.69292 25.98014 26.24801 26.49782
#> 19 17.24579 16.98760 16.74681 16.52224 16.31282 16.11750 15.93535 15.76548
#> 20 20.78334 20.84409 20.90075 20.95359 21.00287 21.04882 21.09168 21.13165
#> 21 17.24579 16.98760 16.74681 16.52224 16.31282 16.11750 15.93535 15.76548
#>      1.6357   1.5254   1.4226   1.3267   1.2373   1.1539   1.0762   1.0036
#> 1  21.15967 21.15967 21.15967 21.15967 21.15967 21.15967 21.23949 21.77292
#> 2  21.15967 21.15967 21.15967 21.15967 21.15967 21.15967 21.18622 21.36323
#> 3  26.67293 26.67293 26.67293 26.67293 26.67293 26.67293 26.66082 26.57595
#> 4  15.64641 15.64641 15.64641 15.64641 15.64641 15.64641 15.70953 16.13444
#> 5  15.64641 15.64641 15.64641 15.64641 15.64641 15.64641 15.68237 15.92558
#> 6  26.67293 26.67293 26.67293 26.67293 26.67293 26.67293 26.47906 25.17820
#> 7  26.67293 26.67293 26.67293 26.67293 26.67293 26.67293 26.48741 25.24246
#> 8  15.64641 15.64641 15.64641 15.64641 15.64641 15.64641 15.57791 15.12227
#> 9  15.64641 15.64641 15.64641 15.64641 15.64641 15.64641 15.64895 15.66852
#> 10 15.64641 15.64641 15.64641 15.64641 15.64641 15.64641 15.63850 15.58819
#> 11 15.64641 15.64641 15.64641 15.64641 15.64641 15.64641 15.29503 12.94691
#> 12 26.67293 26.67293 26.67293 26.67293 26.67293 26.67293 26.80811 27.70862
#> 13 26.67293 26.67293 26.67293 26.67293 26.67293 26.67293 26.76214 27.35516
#> 14 15.64641 15.64641 15.64641 15.64641 15.64641 15.64641 15.69282 16.00591
#> 15 15.64641 15.64641 15.64641 15.64641 15.64641 15.64641 15.71058 16.14247
#> 16 15.64641 15.64641 15.64641 15.64641 15.64641 15.64641 15.62492 15.48376
#> 17 26.67293 26.67293 26.67293 26.67293 26.67293 26.67293 26.74125 27.19450
#> 18 26.67293 26.67293 26.67293 26.67293 26.67293 26.67293 26.69842 26.86515
#> 19 15.64641 15.64641 15.64641 15.64641 15.64641 15.64641 15.76594 16.56823
#> 20 21.15967 21.15967 21.15967 21.15967 21.15967 21.15967 21.20816 21.53193
#> 21 15.64641 15.64641 15.64641 15.64641 15.64641 15.64641 15.68237 15.92558
#>      0.9360   0.8729   0.8141   0.7592   0.7080   0.6603   0.6158   0.5743
#> 1  22.27030 22.38085 22.38094 22.38099 22.35416 22.28126 22.21340 22.14999
#> 2  21.52828 21.56495 21.56496 21.56497 21.53800 21.46489 21.39683 21.33323
#> 3  26.49685 26.47910 26.47892 26.47881 26.50046 26.55953 26.61453 26.66591
#> 4  16.53062 16.61881 16.61901 16.61913 16.65569 16.75436 16.84620 16.93204
#> 5  16.15234 16.20286 16.20302 16.20312 16.19475 16.17172 16.15027 16.13024
#> 6  23.96527 23.69545 23.69501 23.69475 23.69350 23.69103 23.68877 23.68661
#> 7  24.08167 23.82343 23.82301 23.82275 23.82152 23.81909 23.81686 23.81473
#> 8  14.69741 14.60306 14.60308 14.60309 14.61688 14.65422 14.68897 14.72145
#> 9  15.68676 15.69092 15.69104 15.69111 15.70509 15.74271 15.77773 15.81047
#> 10 15.54127 15.53094 15.53105 15.53111 15.54506 15.58264 15.61762 15.65032
#> 11 10.75746 10.27080 10.27042 10.27020 10.26082 10.23622 10.21334 10.19191
#> 12 28.54830 28.73482 28.73484 28.73486 28.73446 28.73333 28.73228 28.73131
#> 13 27.90813 28.03091 28.03087 28.03084 28.05275 28.11224 28.16761 28.21935
#> 14 16.29783 16.36284 16.36302 16.36313 16.39965 16.49825 16.59002 16.67580
#> 15 16.54517 16.63481 16.63501 16.63513 16.67170 16.77037 16.86221 16.94805
#> 16 15.35213 15.32297 15.32305 15.32310 15.35945 15.45777 15.54929 15.63483
#> 17 27.61714 27.71095 27.71088 27.71083 27.73269 27.79209 27.84739 27.89906
#> 18 27.02062 27.05503 27.05490 27.05482 27.05414 27.05257 27.05111 27.04974
#> 19 17.31628 17.48270 17.48298 17.48315 17.47500 17.45230 17.43116 17.41143
#> 20 21.83382 21.90091 21.90095 21.90098 21.82920 21.63459 21.45342 21.28413
#> 21 16.15234 16.20286 16.20302 16.20312 16.10503 15.83881 15.59099 15.35942
#>      0.5356   0.4995   0.4658   0.4344   0.4052   0.3779   0.3524   0.3286
#> 1  22.09085 22.03570 21.98427 21.93630 21.89156 21.84984 21.81094 21.77465
#> 2  21.27391 21.21859 21.16700 21.11889 21.07402 21.03217 20.99315 20.95675
#> 3  26.71382 26.75851 26.80018 26.83904 26.87529 26.90909 26.94061 26.97001
#> 4  17.01209 17.08676 17.15639 17.22133 17.28189 17.33837 17.39104 17.44017
#> 5  16.11156 16.09414 16.07790 16.06275 16.04862 16.03544 16.02316 16.01170
#> 6  23.68459 23.68270 23.68094 23.67930 23.67777 23.67635 23.67502 23.67377
#> 7  23.81274 23.81088 23.80914 23.80752 23.80602 23.80461 23.80330 23.80207
#> 8  14.75174 14.78000 14.80634 14.83091 14.85383 14.87520 14.89513 14.91372
#> 9  15.84100 15.86947 15.89603 15.92079 15.94389 15.96543 15.98552 16.00425
#> 10 15.68081 15.70926 15.73578 15.76052 15.78359 15.80510 15.82517 15.84388
#> 11 10.17193 10.15329 10.13591 10.11970 10.10458 10.09048 10.07733 10.06507
#> 12 28.73039 28.72954 28.72875 28.72801 28.72732 28.72668 28.72608 28.72552
#> 13 28.26761 28.31261 28.35458 28.39372 28.43023 28.46427 28.49602 28.52563
#> 14 16.75580 16.83041 16.89999 16.96488 17.02540 17.08184 17.13448 17.18357
#> 15 17.02811 17.10278 17.17241 17.23735 17.29792 17.35440 17.40708 17.45620
#> 16 15.71460 15.78900 15.85838 15.92309 15.98343 16.03971 16.09220 16.14115
#> 17 27.94724 27.99218 28.03409 28.07317 28.10962 28.14362 28.17532 28.20488
#> 18 27.04846 27.04726 27.04615 27.04511 27.04414 27.04323 27.04239 27.04160
#> 19 17.39304 17.37588 17.35988 17.34496 17.33105 17.31807 17.30597 17.29468
#> 20 21.12625 20.97900 20.84168 20.71361 20.59417 20.48279 20.37891 20.28203
#> 21 15.14346 14.94205 14.75421 14.57903 14.41566 14.26330 14.12121 13.98869
#>      0.3065   0.2858    0.2666    0.2486    0.2319    0.2162    0.2017
#> 1  21.74081 21.73344 21.804244 21.804220 21.804200 21.804185 21.804172
#> 2  20.92281 20.91541 21.119918 21.119862 21.119816 21.119779 21.119749
#> 3  26.99743 27.00340 26.067664 26.067479 26.067329 26.067208 26.067109
#> 4  17.48598 17.49598 17.360479 17.360722 17.360920 17.361080 17.361210
#> 5  16.00101 15.99869 15.875933 15.875999 15.876053 15.876096 15.876132
#> 6  23.67262 23.67234 23.329477 23.329125 23.328840 23.328609 23.328421
#> 7  23.80093 23.80066 25.752235 25.752216 25.752201 25.752189 25.752179
#> 8  14.93105 14.93483 14.892111 14.892165 14.892209 14.892244 14.892272
#> 9  16.02172 16.02554 16.538160 16.538361 16.538523 16.538655 16.538762
#> 10 15.86133 15.86514 16.628864 16.629091 16.629274 16.629423 16.629543
#> 11 10.05363 10.05112  9.279661  9.279263  9.278941  9.278679  9.278467
#> 12 28.72500 28.72489 29.089675 29.089740 29.089792 29.089834 29.089869
#> 13 28.55324 28.55926 29.194295 29.194430 29.194540 29.194629 29.194701
#> 14 17.22935 17.23934 16.893399 16.893597 16.893757 16.893887 16.893993
#> 15 17.50202 17.51202 17.599390 17.599667 17.599891 17.600072 17.600219
#> 16 16.18680 16.19675 15.607612 15.607710 15.607789 15.607853 15.607906
#> 17 28.23246 28.23847 27.980804 27.980792 27.980783 27.980775 27.980769
#> 18 27.04087 27.04070 25.376904 25.376576 25.376310 25.376094 25.375919
#> 19 17.28415 17.28187 16.591695 16.591762 16.591815 16.591859 16.591894
#> 20 20.19168 20.17199 20.402476 20.402292 20.402142 20.402021 20.401923
#> 21 13.86511 13.83817 14.915002 14.914932 14.914875 14.914829 14.914792
#>       0.1881    0.1754    0.1636    0.1525    0.1423    0.1327    0.1237
#> 1  21.804162 21.777029 21.746189 21.603024 21.603052 21.603077 21.603099
#> 2  21.119725 21.086209 21.048396 21.024409 21.024396 21.024384 21.024374
#> 3  26.067029 26.126687 26.197155 25.999331 25.999134 25.998973 25.998840
#> 4  17.361315 17.401755 17.445445 17.496121 17.496294 17.496440 17.496561
#> 5  15.876160 15.874793 15.872511 15.700774 15.700863 15.700939 15.701004
#> 6  23.328269 23.339684 23.356648 23.151643 23.151342 23.151092 23.150885
#> 7  25.752171 25.663365 25.561641 25.530087 25.530160 25.530218 25.530264
#> 8  14.892296 14.906862 14.922972 14.985574 14.985599 14.985619 14.985635
#> 9  16.538849 16.529855 16.517311 16.380932 16.381144 16.381322 16.381472
#> 10 16.629641 16.609565 16.584019 16.458723 16.458967 16.459172 16.459343
#> 11  9.278295  9.301374  9.332242  9.824616  9.824155  9.823765  9.823437
#> 12 29.089897 29.073700 29.054403 29.747392 29.747360 29.747325 29.747291
#> 13 29.194759 29.186772 29.176088 29.164478 29.164609 29.164718 29.164809
#> 14 16.894078 16.943549 16.998109 16.559664 16.559862 16.560033 16.560180
#> 15 17.600338 17.631046 17.663208 17.699228 17.699443 17.699622 17.699772
#> 16 15.607948 15.667365 15.734434 15.913799 15.913802 15.913803 15.913804
#> 17 27.980764 28.011591 28.047067 27.939873 27.939848 27.939828 27.939813
#> 18 25.375777 25.447155 25.532686 25.904213 25.903788 25.903430 25.903131
#> 19 16.591922 16.616205 16.643366 17.703455 17.703329 17.703215 17.703113
#> 20 20.401843 20.326438 20.242019 19.407306 19.407344 19.407382 19.407420
#> 21 14.914762 14.779002 14.624094 14.105357 14.105510 14.105641 14.105753
#>      0.1154   0.1076   0.1004   0.0936   0.0873   0.0814   0.0759   0.0708
#> 1  21.60312 22.06103 22.06102 22.06100 22.06098 22.06095 22.06092 22.29668
#> 2  21.02437 21.40705 21.40712 21.40717 21.40721 21.40722 21.40723 21.69597
#> 3  25.99873 25.86024 25.85998 25.85976 25.85956 25.85938 25.85922 25.57163
#> 4  17.49666 17.39113 17.39106 17.39102 17.39100 17.39099 17.39100 17.38449
#> 5  15.70106 16.36953 16.36945 16.36937 16.36929 16.36922 16.36915 15.77288
#> 6  23.15071 23.04734 23.04721 23.04709 23.04696 23.04684 23.04672 23.19183
#> 7  25.53030 25.20829 25.20853 25.20874 25.20894 25.20913 25.20930 25.11639
#> 8  14.98565 15.07897 15.07897 15.07897 15.07898 15.07897 15.07897 15.14348
#> 9  16.38160 16.46758 16.46754 16.46752 16.46752 16.46753 16.46755 16.47579
#> 10 16.45949 16.49772 16.49771 16.49773 16.49776 16.49781 16.49786 16.52089
#> 11  9.82316 10.08272 10.08295 10.08308 10.08315 10.08317 10.08315 10.09035
#> 12 29.74726 30.47270 30.47315 30.47349 30.47376 30.47395 30.47409 30.43913
#> 13 29.16488 29.07750 29.07752 29.07756 29.07761 29.07766 29.07771 28.97308
#> 14 16.56031 16.28290 16.28253 16.28225 16.28204 16.28188 16.28177 16.48977
#> 15 17.69990 17.56276 17.56271 17.56268 17.56268 17.56269 17.56272 17.77247
#> 16 15.91380 15.75326 15.75320 15.75316 15.75313 15.75310 15.75308 15.81587
#> 17 27.93980 27.89506 27.89494 27.89483 27.89474 27.89466 27.89459 27.78010
#> 18 25.90288 25.47890 25.47890 25.47887 25.47882 25.47875 25.47868 25.63079
#> 19 17.70302 16.83187 16.83234 16.83274 16.83308 16.83337 16.83362 16.94770
#> 20 19.40746 19.35690 19.35660 19.35636 19.35618 19.35603 19.35592 19.51068
#> 21 14.10585 14.11654 14.11656 14.11659 14.11664 14.11670 14.11677 13.68004
#>      0.0660   0.0616   0.0574   0.0536   0.0500   0.0466   0.0434   0.0405
#> 1  22.29653 22.23086 22.23078 22.23072 22.23067 22.45896 22.45886 22.45876
#> 2  21.69579 21.63419 21.63406 21.63395 21.63386 21.78790 21.78779 21.78769
#> 3  25.57155 25.73546 25.73550 25.73550 25.73549 25.61197 25.61222 25.61245
#> 4  17.38468 17.74405 17.74387 17.74371 17.74355 17.79355 17.79336 17.79316
#> 5  15.77280 15.25278 15.25300 15.25320 15.25339 15.62701 15.62738 15.62773
#> 6  23.19160 22.98026 22.98077 22.98125 22.98170 22.92597 22.92613 22.92629
#> 7  25.11663 25.24864 25.24865 25.24866 25.24867 25.21506 25.21488 25.21470
#> 8  15.14345 14.48232 14.48241 14.48251 14.48262 14.29424 14.29451 14.29479
#> 9  16.47598 16.06422 16.06419 16.06418 16.06417 15.99611 15.99618 15.99626
#> 10 16.52114 16.18828 16.18818 16.18810 16.18804 16.11195 16.11196 16.11197
#> 11 10.08973 10.35314 10.35311 10.35309 10.35305 10.49053 10.49064 10.49074
#> 12 30.43873 29.99740 29.99733 29.99729 29.99728 30.32557 30.32570 30.32582
#> 13 28.97326 29.26704 29.26675 29.26648 29.26621 29.26365 29.26351 29.26337
#> 14 16.49010 16.85772 16.85763 16.85753 16.85744 16.76874 16.76850 16.76826
#> 15 17.77270 17.83453 17.83438 17.83425 17.83414 17.74740 17.74723 17.74706
#> 16 15.81588 16.39040 16.39023 16.39006 16.38989 16.42770 16.42749 16.42727
#> 17 27.78011 27.93715 27.93706 27.93697 27.93687 27.89869 27.89874 27.89878
#> 18 25.63045 25.63720 25.63757 25.63791 25.63823 25.35435 25.35456 25.35477
#> 19 16.94773 17.41936 17.41916 17.41897 17.41880 16.98691 16.98674 16.98658
#> 20 19.51086 19.42303 19.42332 19.42359 19.42383 19.45304 19.45296 19.45288
#> 21 13.68033 13.62195 13.62203 13.62209 13.62212 13.76067 13.76067 13.76067
#>      0.0378   0.0352   0.0329   0.0306   0.0286   0.0267   0.0249   0.0232
#> 1  22.45866 22.38964 22.38970 22.38976 22.38982 22.38988 22.38993 22.38998
#> 2  21.78760 21.69781 21.69782 21.69783 21.69785 21.69786 21.69787 21.69788
#> 3  25.61266 25.66817 25.66817 25.66819 25.66820 25.66822 25.66824 25.66826
#> 4  17.79296 17.73991 17.74009 17.74027 17.74044 17.74061 17.74077 17.74093
#> 5  15.62807 15.53845 15.53848 15.53853 15.53858 15.53863 15.53869 15.53876
#> 6  22.92645 23.17675 23.17660 23.17645 23.17630 23.17615 23.17600 23.17585
#> 7  25.21453 25.07133 25.07141 25.07149 25.07156 25.07164 25.07171 25.07179
#> 8  14.29507 14.34394 14.34361 14.34329 14.34298 14.34268 14.34239 14.34210
#> 9  15.99634 15.97562 15.97548 15.97535 15.97522 15.97509 15.97497 15.97485
#> 10 16.11199 16.05737 16.05727 16.05716 16.05706 16.05697 16.05687 16.05678
#> 11 10.49083 10.48950 10.48953 10.48956 10.48959 10.48962 10.48965 10.48969
#> 12 30.32594 30.47858 30.47843 30.47829 30.47815 30.47801 30.47787 30.47773
#> 13 29.26323 29.23321 29.23333 29.23345 29.23356 29.23367 29.23377 29.23388
#> 14 16.76803 16.83202 16.83216 16.83230 16.83244 16.83256 16.83268 16.83279
#> 15 17.74690 17.77366 17.77365 17.77365 17.77364 17.77363 17.77362 17.77360
#> 16 16.42706 16.40325 16.40346 16.40367 16.40386 16.40406 16.40424 16.40442
#> 17 27.89881 27.97434 27.97437 27.97441 27.97444 27.97447 27.97450 27.97453
#> 18 25.35497 25.05880 25.05869 25.05859 25.05849 25.05841 25.05834 25.05827
#> 19 16.98643 17.11278 17.11267 17.11256 17.11244 17.11232 17.11220 17.11208
#> 20 19.45281 19.48720 19.48726 19.48731 19.48735 19.48740 19.48744 19.48748
#> 21 13.76066 13.79767 13.79778 13.79790 13.79801 13.79812 13.79822 13.79833
#>      0.0216   0.0202   0.0188   0.0175   0.0164   0.0153   0.0142   0.0133
#> 1  22.39003 22.39008 22.39013 22.39018 22.39022 22.39027 22.39031 22.39035
#> 2  21.69790 21.69791 21.69792 21.69793 21.69794 21.69795 21.69796 21.69796
#> 3  25.66829 25.66831 25.66834 25.66837 25.66841 25.66844 25.66848 25.66851
#> 4  17.74108 17.74123 17.74137 17.74151 17.74165 17.74177 17.74190 17.74202
#> 5  15.53883 15.53890 15.53898 15.53906 15.53915 15.53923 15.53933 15.53942
#> 6  23.17570 23.17556 23.17541 23.17527 23.17512 23.17498 23.17484 23.17471
#> 7  25.07186 25.07193 25.07200 25.07207 25.07214 25.07220 25.07227 25.07233
#> 8  14.34182 14.34155 14.34129 14.34104 14.34079 14.34055 14.34032 14.34010
#> 9  15.97474 15.97463 15.97452 15.97442 15.97431 15.97422 15.97412 15.97403
#> 10 16.05669 16.05661 16.05652 16.05644 16.05637 16.05629 16.05622 16.05615
#> 11 10.48972 10.48976 10.48979 10.48983 10.48987 10.48991 10.48994 10.48998
#> 12 30.47759 30.47746 30.47733 30.47720 30.47707 30.47695 30.47682 30.47670
#> 13 29.23398 29.23408 29.23417 29.23427 29.23436 29.23444 29.23453 29.23461
#> 14 16.83290 16.83300 16.83309 16.83318 16.83326 16.83334 16.83341 16.83348
#> 15 17.77359 17.77357 17.77354 17.77352 17.77349 17.77346 17.77343 17.77340
#> 16 16.40460 16.40477 16.40493 16.40509 16.40524 16.40539 16.40553 16.40566
#> 17 27.97456 27.97459 27.97461 27.97464 27.97467 27.97469 27.97472 27.97474
#> 18 25.05821 25.05816 25.05811 25.05807 25.05804 25.05801 25.05799 25.05798
#> 19 17.11196 17.11183 17.11171 17.11158 17.11145 17.11133 17.11120 17.11107
#> 20 19.48752 19.48755 19.48758 19.48761 19.48764 19.48766 19.48768 19.48771
#> 21 13.79843 13.79853 13.79863 13.79873 13.79882 13.79892 13.79901 13.79910
#>      0.0124   0.0115   0.0108   0.0100   0.0094   0.0087   0.0081   0.0076
#> 1  22.39039 22.39042 22.39046 22.39050 22.39053 22.39056 22.39059 22.39062
#> 2  21.69797 21.69798 21.69799 21.69800 21.69800 21.69801 21.69801 21.69802
#> 3  25.66855 25.66859 25.66863 25.66867 25.66872 25.66876 25.66880 25.66885
#> 4  17.74214 17.74225 17.74236 17.74246 17.74256 17.74265 17.74275 17.74283
#> 5  15.53951 15.53961 15.53971 15.53982 15.53992 15.54002 15.54013 15.54024
#> 6  23.17457 23.17443 23.17430 23.17417 23.17404 23.17392 23.17379 23.17367
#> 7  25.07240 25.07246 25.07252 25.07258 25.07264 25.07269 25.07275 25.07280
#> 8  14.33988 14.33967 14.33947 14.33927 14.33908 14.33890 14.33873 14.33856
#> 9  15.97394 15.97385 15.97377 15.97369 15.97361 15.97354 15.97347 15.97340
#> 10 16.05608 16.05601 16.05595 16.05589 16.05583 16.05577 16.05572 16.05566
#> 11 10.49002 10.49006 10.49010 10.49014 10.49018 10.49022 10.49026 10.49030
#> 12 30.47658 30.47646 30.47635 30.47624 30.47613 30.47602 30.47591 30.47581
#> 13 29.23469 29.23477 29.23484 29.23491 29.23498 29.23505 29.23511 29.23518
#> 14 16.83354 16.83360 16.83365 16.83370 16.83374 16.83378 16.83382 16.83385
#> 15 17.77337 17.77333 17.77330 17.77326 17.77322 17.77318 17.77314 17.77310
#> 16 16.40580 16.40592 16.40604 16.40616 16.40627 16.40638 16.40648 16.40658
#> 17 27.97476 27.97478 27.97481 27.97483 27.97485 27.97487 27.97489 27.97490
#> 18 25.05797 25.05797 25.05797 25.05797 25.05798 25.05800 25.05802 25.05804
#> 19 17.11094 17.11081 17.11068 17.11056 17.11043 17.11030 17.11018 17.11005
#> 20 19.48772 19.48774 19.48776 19.48777 19.48778 19.48779 19.48780 19.48781
#> 21 13.79918 13.79927 13.79935 13.79943 13.79950 13.79958 13.79965 13.79973
#>      0.0071   0.0066   0.0062   0.0057   0.0054   0.0050
#> 1  22.39065 22.39068 22.39070 22.39073 22.39075 22.39077
#> 2  21.69803 21.69803 21.69804 21.69804 21.69804 21.69805
#> 3  25.66889 25.66894 25.66898 25.66903 25.66908 25.66912
#> 4  17.74292 17.74300 17.74308 17.74315 17.74322 17.74329
#> 5  15.54034 15.54045 15.54056 15.54067 15.54078 15.54089
#> 6  23.17355 23.17343 23.17331 23.17320 23.17309 23.17298
#> 7  25.07286 25.07291 25.07296 25.07301 25.07306 25.07310
#> 8  14.33840 14.33824 14.33809 14.33795 14.33781 14.33768
#> 9  15.97333 15.97327 15.97321 15.97315 15.97309 15.97304
#> 10 16.05561 16.05556 16.05552 16.05547 16.05543 16.05539
#> 11 10.49034 10.49038 10.49042 10.49046 10.49050 10.49054
#> 12 30.47571 30.47561 30.47551 30.47542 30.47533 30.47524
#> 13 29.23524 29.23530 29.23535 29.23540 29.23546 29.23551
#> 14 16.83388 16.83391 16.83393 16.83394 16.83396 16.83397
#> 15 17.77306 17.77302 17.77298 17.77293 17.77289 17.77284
#> 16 16.40667 16.40676 16.40685 16.40693 16.40700 16.40708
#> 17 27.97492 27.97494 27.97495 27.97497 27.97498 27.97500
#> 18 25.05806 25.05809 25.05813 25.05816 25.05820 25.05824
#> 19 17.10993 17.10981 17.10968 17.10956 17.10945 17.10933
#> 20 19.48781 19.48782 19.48782 19.48782 19.48782 19.48782
#> 21 13.79979 13.79986 13.79993 13.79999 13.80005 13.80011
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 22.8 18.7 14.3 24.4 22.8 16.4 17.3 15.2 10.4 30.4 33.9 15.5 15.2
#> [16] 19.2 27.3 26.0 15.8 19.7 15.0
#> 
#> $X
#>               am        carb        cyl       disp        drat       gear
#>  [1,]  1.0488088  0.62526247 -0.2102353 -0.6242211  0.51513377  0.3175003
#>  [2,]  1.0488088  0.62526247 -0.2102353 -0.6242211  0.51513377  0.3175003
#>  [3,]  1.0488088 -1.16526187 -1.3139708 -1.0703267  0.42558237  0.3175003
#>  [4,] -0.9534626 -0.56842042  0.8935002  1.0915698 -0.82813723 -1.0160010
#>  [5,] -0.9534626  0.62526247  0.8935002  1.0915698 -0.72067555 -1.0160010
#>  [6,] -0.9534626 -0.56842042 -1.3139708 -0.7383212  0.13901789  0.3175003
#>  [7,] -0.9534626 -0.56842042 -1.3139708 -0.7889370  0.55095433  0.3175003
#>  [8,] -0.9534626  0.02842102  0.8935002  0.3692219 -0.97141947 -1.0160010
#>  [9,] -0.9534626  0.02842102  0.8935002  0.3692219 -0.97141947 -1.0160010
#> [10,] -0.9534626  0.02842102  0.8935002  0.3692219 -0.97141947 -1.0160010
#> [11,] -0.9534626  0.62526247  0.8935002  1.9494653 -1.09679143 -1.0160010
#> [12,]  1.0488088 -0.56842042 -1.3139708 -1.3474269  2.35989261  0.3175003
#> [13,]  1.0488088 -1.16526187 -1.3139708 -1.3868901  1.08826273  0.3175003
#> [14,] -0.9534626 -0.56842042  0.8935002  0.7312537 -1.52663815 -1.0160010
#> [15,] -0.9534626 -0.56842042  0.8935002  0.6111484 -0.82813723 -1.0160010
#> [16,] -0.9534626 -0.56842042  0.8935002  1.4347280 -0.95350919 -1.0160010
#> [17,]  1.0488088 -1.16526187 -1.3139708 -1.3191164  0.83751881  0.3175003
#> [18,]  1.0488088 -0.56842042 -1.3139708 -0.9648056  1.46437861  1.6510017
#> [19,]  1.0488088  0.62526247  0.8935002  1.0143593  1.08826273  1.6510017
#> [20,]  1.0488088  1.81894535 -0.2102353 -0.7529054  0.01364593  1.6510017
#> [21,]  1.0488088  3.01262824  0.8935002  0.5854115 -0.12963631  1.6510017
#>                hp        qsec         vs           wt
#>  [1,] -0.56551418 -0.58474749 -0.6324555 -0.605669123
#>  [2,] -0.56551418 -0.28380086 -0.6324555 -0.310388883
#>  [3,] -0.80081553  0.57067260  1.5811388 -0.953057641
#>  [4,]  0.33416747 -0.28380086 -0.6324555  0.343859491
#>  [5,]  1.30305539 -0.91793840 -0.6324555  0.494394516
#>  [6,] -1.22989446  1.31766512  1.5811388  0.054369060
#>  [7,] -0.77313302  2.87613873  1.5811388  0.008050591
#>  [8,]  0.40337375 -0.07958708 -0.6324555  1.073375378
#>  [9,]  0.40337375  0.02789386 -0.6324555  0.679668392
#> [10,]  0.40337375  0.24285574 -0.6324555  0.737566478
#> [11,]  0.88781771  0.14612289 -0.6324555  2.641255554
#> [12,] -1.36830703  0.52230618  1.5811388 -1.769420657
#> [13,] -1.18837070  1.26392466  1.5811388 -1.514669078
#> [14,] -0.01186393 -0.36441156 -0.6324555  0.436496429
#> [15,] -0.01186393 -0.13332755 -0.6324555  0.338069683
#> [16,]  0.33416747 -0.26767872 -0.6324555  0.812833990
#> [17,] -1.17452944  0.72651996  1.5811388 -1.398872905
#> [18,] -0.82849804 -0.45577036 -0.6324555 -1.161490751
#> [19,]  1.56603925 -1.63806068 -0.6324555  0.031209826
#> [20,]  0.33416747 -1.10065599 -0.6324555 -0.431974865
#> [21,]  2.54876843 -1.58432021 -0.6324555  0.494394516
#> attr(,"center")
#>  [1]   0.4761905   2.9523810   6.3809524 232.7619048   3.6123810   3.7619048
#>  [7] 150.8571429  17.5480952   0.2857143   3.1430476
#> attr(,"scale")
#>  [1]   0.4994328   1.6754869   1.8120283 116.5643190   0.5583386   0.7499055
#>  [7]  72.2477785   1.8607951   0.4517540   0.8635864
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 37
#> 
#> $lambda.min
#> [1] 0.4051673
#> 
#> $null.dev
#> [1] 32.55896
#> 
#> $Bias
#> [1] 4.35258
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 8.662935