Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 41.323681 38.845344 34.346906 28.927057 24.440954 20.751622 17.740963
#>   [8] 15.307225 13.362809 11.832368 10.651149  9.763564  9.121946  8.685472
#>  [15]  8.419228  8.292618  8.215525  8.143820  8.184789  8.261025  8.321577
#>  [22]  8.300225  8.305407  8.332068  8.375883  8.387920  8.304843  8.245279
#>  [29]  8.133704  7.908868  7.723483  7.571349  7.591682  7.941579  8.575953
#>  [36]  9.172847  9.340437  9.252130  9.063546  8.862116  8.733239  8.545965
#>  [43]  8.401464  8.347229  8.316222  8.938463  8.938555  8.892534  9.417455
#>  [50] 11.031572 11.993182 12.270620 13.932174 14.376655 13.417186 13.711034
#>  [57] 14.296498 14.233194 14.085879 14.086151 14.059120 14.318290 16.485858
#>  [64] 16.555753 17.302262 17.543660 17.544629 17.569235 17.564150 17.780873
#>  [71] 17.781274 17.770310 17.770585 20.223795 19.601116 19.601881 20.420193
#>  [78] 20.422328 20.424404 20.426422 20.428384 20.430292 20.432147 20.433951
#>  [85] 20.435706 20.636465 20.638848 20.626036 20.628108 20.630127 20.632094
#>  [92] 20.634011 20.635877 20.637696 20.617369 20.619153 20.620891 20.622584
#>  [99] 20.624233 20.625839
#> 
#> $cvse
#>   [1] 11.809984 11.554768 10.582991  9.157552  8.005468  7.072131  6.310162
#>   [8]  5.679763  5.148839  4.692725  4.293499  3.939037  3.621989  3.338801
#>  [15]  3.088838  2.873589  2.699770  2.564161  2.535782  2.541055  2.540679
#>  [22]  2.533515  2.531213  2.533300  2.539313  2.526565  2.466357  2.418439
#>  [29]  2.365653  2.298703  2.251362  2.219944  2.177052  2.127861  2.129408
#>  [36]  2.171100  2.164104  2.123488  2.071017  2.021881  2.002677  2.002015
#>  [43]  2.006561  2.010162  2.011096  2.004652  2.004659  1.982447  2.309605
#>  [50]  2.640024  3.047939  3.052727  3.485131  3.864985  3.631000  3.592154
#>  [57]  3.648380  3.640306  3.678598  3.678634  3.667624  3.643353  4.890697
#>  [64]  4.898592  5.096385  5.177400  5.177998  5.173862  5.174566  5.192319
#>  [71]  5.192730  5.195858  5.196189  6.426653  6.295863  6.296572  6.560534
#>  [78]  6.561719  6.562875  6.564003  6.565105  6.566181  6.567231  6.568257
#>  [85]  6.569259  6.642602  6.643615  6.645921  6.646898  6.647851  6.648779
#>  [92]  6.649685  6.650567  6.651428  6.653743  6.654558  6.655353  6.656130
#>  [99]  6.656887  6.657627
#> 
#> $fold
#>  [1]  5  2  9  3  4  8  1  5  6  9  7  8 10  2  4  6 10  1  7  3  1
#> 
#> $lambda
#>   [1] 5.757029246 5.369024743 5.007170445 4.669703915 4.354981500 4.061470323
#>   [7] 3.787740816 3.532459762 3.294383797 3.072353355 2.865287022 2.672176266
#>  [13] 2.492080529 2.324122642 2.167484555 2.021403350 1.885167530 1.758113548
#>  [19] 1.639622579 1.529117505 1.426060103 1.329948424 1.240314352 1.156721316
#>  [25] 1.078762170 1.006057211 0.938252322 0.875017256 0.816044021 0.761045385
#>  [31] 0.709753473 0.661918465 0.617307376 0.575702925 0.536902474 0.500717044
#>  [37] 0.466970392 0.435498150 0.406147032 0.378774082 0.353245976 0.329438380
#>  [43] 0.307235335 0.286528702 0.267217627 0.249208053 0.232412264 0.216748455
#>  [49] 0.202140335 0.188516753 0.175811355 0.163962258 0.152911750 0.142606010
#>  [55] 0.132994842 0.124031435 0.115672132 0.107876217 0.100605721 0.093825232
#>  [61] 0.087501726 0.081604402 0.076104539 0.070975347 0.066191846 0.061730738
#>  [67] 0.057570292 0.053690247 0.050071704 0.046697039 0.043549815 0.040614703
#>  [73] 0.037877408 0.035324598 0.032943838 0.030723534 0.028652870 0.026721763
#>  [79] 0.024920805 0.023241226 0.021674846 0.020214034 0.018851675 0.017581135
#>  [85] 0.016396226 0.015291175 0.014260601 0.013299484 0.012403144 0.011567213
#>  [91] 0.010787622 0.010060572 0.009382523 0.008750173 0.008160440 0.007610454
#>  [97] 0.007097535 0.006619185 0.006173074 0.005757029
#> 
#> $fit
#> $beta
#>              5.7570     5.3690    5.0072    4.6697    4.3550    4.0615   3.7877
#> (Intercept) 19.9381 21.7812018 23.500089 25.103129 26.598130 27.992372 29.29265
#> am           0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> carb         0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> cyl          0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> disp         0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> drat         0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> gear         0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> hp           0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> qsec         0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> vs           0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> wt           0.0000 -0.5742021 -1.109705 -1.609116 -2.074869 -2.509232 -2.91432
#>                3.5325    3.2944    3.0724    2.8653    2.6722    2.4921
#> (Intercept) 30.505289 31.636202 32.690895 33.674506 34.591824 35.447318
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -3.292107 -3.644432 -3.973012 -4.279446 -4.565228 -4.831749
#>                2.3241    2.1675    2.0214    1.8852    1.7581    1.6396
#> (Intercept) 36.245155 36.989220 37.683138 38.169530 38.169530 38.169530
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -5.080307 -5.312113 -5.528297 -5.679828 -5.679828 -5.679828
#>                1.5291      1.4261     1.3299     1.2403     1.1567     1.0788
#> (Intercept) 38.169530 37.85562998 35.5811849 33.4600751 31.4819209 29.6370877
#> am           0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.000000  0.01633708  0.1347115  0.2451056  0.3480595  0.4440746
#> vs           0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.679828 -5.67364396 -5.6288391 -5.5870548 -5.5480866 -5.5117447
#>                 1.0061     0.9383     0.8750     0.8160     0.7610     0.7098
#> (Intercept) 27.9165901 26.3120483 24.8156472 23.4200986 22.1186052 20.9048282
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.5336187  0.6171278  0.6950086  0.7676405  0.8353773  0.8985489
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.4778522 -5.4462439 -5.4167659 -5.3892746 -5.3636362 -5.3397256
#>                 0.6619    0.6173    0.5757    0.5369    0.5007    0.4670
#> (Intercept) 19.7728557 18.717174 17.732642 16.814464 16.416552 15.095886
#> am           0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.0000000  0.000000  0.000000  0.000000  0.000000  0.180787
#> hp           0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.9574628  1.012406  1.063647  1.111434  1.132143  1.155239
#> vs           0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -5.3174266 -5.296630 -5.277236 -5.259148 -5.251310 -5.170527
#>                 0.4355     0.4061    0.3788     0.3532    0.3294    0.3072
#> (Intercept) 13.5544706 12.1139865 10.770560  9.5229532  8.354221  7.264191
#> am           0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.000000
#> carb         0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.000000
#> cyl          0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.000000
#> disp         0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.000000
#> drat         0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.000000
#> gear         0.3916378  0.5886394  0.772367  0.9430664  1.102898  1.251967
#> hp           0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.000000
#> qsec         1.1822335  1.2074715  1.231009  1.2528488  1.273327  1.292426
#> vs           0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.000000
#> wt          -5.0762876 -4.9882321 -4.906110 -4.8298218 -4.758380 -4.691749
#>                0.2865    0.2672    0.2492    0.2324    0.2167    0.2021
#> (Intercept)  6.631208  6.628380  6.626973  6.626274  6.625926 -7.871343
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         1.338493  1.338793  1.338942  1.339017  1.339054  2.375017
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         1.303527  1.303598  1.303634  1.303651  1.303660  2.076257
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000 -3.258664
#> wt          -4.653067 -4.652921 -4.652848 -4.652812 -4.652793 -5.089477
#>                0.1885    0.1758    0.1640    0.1529    0.1426    0.1330
#> (Intercept) -7.878282 -7.884215 -7.889289 -7.893628 -7.897337 -7.900510
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         2.375532  2.375972  2.376349  2.376671  2.376946  2.377182
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         2.076579  2.076855  2.077090  2.077292  2.077464  2.077611
#> vs          -3.259726 -3.260633 -3.261409 -3.262073 -3.262640 -3.263126
#> wt          -5.089522 -5.089560 -5.089593 -5.089621 -5.089645 -5.089665
#>                0.1240    0.1157      0.1079      0.1006      0.0938      0.0875
#> (Intercept) -7.903222 -7.905542  1.84705742  1.84463348  1.84230846  1.84007854
#> am           0.000000  0.000000  2.45272874  2.45306931  2.45339480  2.45370586
#> carb         0.000000  0.000000  1.86803669  1.86832264  1.86859599  1.86885730
#> cyl          0.000000  0.000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.000000  0.000000  0.02201450  0.02201615  0.02201773  0.02201923
#> drat         0.000000  0.000000 -1.59410358 -1.59447265 -1.59482553 -1.59516291
#> gear         2.377383  2.377556  0.90083995  0.90063832  0.90044575  0.90026184
#> hp           0.000000  0.000000 -0.03222834 -0.03222376 -0.03221937 -0.03221516
#> qsec         2.077737  2.077845  2.22531575  2.22556638  2.22580638  2.22603619
#> vs          -3.263540 -3.263895 -2.31404387 -2.31400276 -2.31396391 -2.31392720
#> wt          -5.089683 -5.089698 -7.55000304 -7.55059831 -7.55116772 -7.55171238
#>                  0.0816      0.0761      0.0710      0.0662      0.0617
#> (Intercept)  1.83794006  1.83588946  1.83392333  1.83203835  1.83023134
#> am           2.45400313  2.45428720  2.45455866  2.45481806  2.45506593
#> carb         1.86910708  1.86934583  1.86957403  1.86979214  1.87000061
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.02202067  0.02202205  0.02202336  0.02202461  0.02202581
#> drat        -1.59548547 -1.59579383 -1.59608862 -1.59637042 -1.59663980
#> gear         0.90008620  0.89991846  0.89975827  0.89960529  0.89945921
#> hp          -0.03221113 -0.03220726 -0.03220356 -0.03220001 -0.03219661
#> qsec         2.22625623  2.22646690  2.22666858  2.22686165  2.22704647
#> vs          -2.31389250 -2.31385972 -2.31382874 -2.31379946 -2.31377180
#> wt          -7.55223333 -7.55273158 -7.55320811 -7.55366383 -7.55409964
#>                  0.0576      0.0537      0.0501      0.0467      0.0435
#> (Intercept)  1.82849924  1.82683908  1.82524801  8.15655283  8.15497616
#> am           2.45530276  2.45552906  2.45574527  2.17911409  2.17863676
#> carb         1.87019984  1.87039025  1.87057221  2.14471599  2.14470508
#> cyl          0.00000000  0.00000000  0.00000000 -0.54933067 -0.54947613
#> disp         0.02202695  0.02202804  0.02202908  0.02426828  0.02426794
#> drat        -1.59689729 -1.59714340 -1.59737864 -1.91302103 -1.91283556
#> gear         0.89931971  0.89918650  0.89905930  0.73265560  0.73304005
#> hp          -0.03219335 -0.03219023 -0.03218724 -0.03195256 -0.03195525
#> qsec         2.22722336  2.22739268  2.22755472  2.17875985  2.17883750
#> vs          -2.31374565 -2.31372094 -2.31369759 -3.25192234 -3.25301555
#> wt          -7.55451639 -7.55491489 -7.55529593 -7.85753026 -7.85745014
#>                  0.0406      0.0379      0.0353      0.0329      0.0307
#> (Intercept)  8.15341480  8.15187583  8.15036522  8.14888800  8.14744831
#> am           2.17819208  2.17777833  2.17739387  2.17703708  2.17670641
#> carb         2.14468759  2.14466456  2.14463693  2.14460552  2.14457110
#> cyl         -0.54960214 -0.54971052 -0.54980297 -0.54988104 -0.54994617
#> disp         0.02426763  0.02426737  0.02426714  0.02426694  0.02426678
#> drat        -1.91265501 -1.91247981 -1.91231032 -1.91214683 -1.91198953
#> gear         0.73340344  0.73374645  0.73406978  0.73437415  0.73466029
#> hp          -0.03195780 -0.03196023 -0.03196254 -0.03196472 -0.03196680
#> qsec         2.17890862  2.17897366  2.17903308  2.17908727  2.17913664
#> vs          -3.25400949 -3.25491152 -3.25572856 -3.25646707 -3.25713312
#> wt          -7.85736998 -7.85729036 -7.85721178 -7.85713466 -7.85705939
#>                  0.0287      0.0267      0.0249      0.0232      0.0217
#> (Intercept)  8.14604953  8.14469435  8.14338485  8.14212257  8.14090857
#> am           2.17640037  2.17611754  2.17585655  2.17561607  2.17539486
#> carb         2.14453432  2.14449578  2.14445600  2.14441545  2.14437453
#> cyl         -0.54999969 -0.55004280 -0.55007661 -0.55010216 -0.55012038
#> disp         0.02426664  0.02426654  0.02426645  0.02426639  0.02426636
#> drat        -1.91183858 -1.91169407 -1.91155607 -1.91142457 -1.91129954
#> gear         0.73492894  0.73518083  0.73541669  0.73563723  0.73584318
#> hp          -0.03196876 -0.03197062 -0.03197238 -0.03197404 -0.03197561
#> qsec         2.17918156  2.17922237  2.17925940  2.17929296  2.17932332
#> vs          -3.25773239 -3.25827019 -3.25875147 -3.25918086 -3.25956267
#> wt          -7.85698626 -7.85691555 -7.85684747 -7.85678218 -7.85671982
#>                  0.0202      0.0189      0.0176      0.0164      0.0153
#> (Intercept)  8.13974350  8.13862764  8.13756096  8.13654314  8.13557364
#> am           2.17519170  2.17500547  2.17483506  2.17467945  2.17453765
#> carb         2.14433359  2.14429294  2.14425286  2.14421355  2.14417522
#> cyl         -0.55013212 -0.55013816 -0.55013921 -0.55013593 -0.55012890
#> disp         0.02426634  0.02426634  0.02426636  0.02426639  0.02426644
#> drat        -1.91118093 -1.91106865 -1.91096258 -1.91086260 -1.91076856
#> gear         0.73603523  0.73621407  0.73638036  0.73653475  0.73667789
#> hp          -0.03197709 -0.03197848 -0.03197980 -0.03198103 -0.03198220
#> qsec         2.17935078  2.17937556  2.17939791  2.17941803  2.17943614
#> vs          -3.25990092 -3.26019933 -3.26046139 -3.26069031 -3.26088908
#> wt          -7.85666050 -7.85660428 -7.85655120 -7.85650129 -7.85645454
#>                  0.0143      0.0133      0.0124      0.0116      0.0108
#> (Intercept)  8.13465171  8.13377644  8.13294677  8.13216152  8.13141943
#> am           2.17440872  2.17429178  2.17418599  2.17409056  2.17400475
#> carb         2.14413803  2.14410211  2.14406756  2.14403448  2.14400291
#> cyl         -0.55011865 -0.55010567 -0.55009040 -0.55007323 -0.55005450
#> disp         0.02426650  0.02426657  0.02426666  0.02426675  0.02426685
#> drat        -1.91068030 -1.91059763 -1.91052039 -1.91044837 -1.91038139
#> gear         0.73681037  0.73693280  0.73704574  0.73714974  0.73724533
#> hp          -0.03198329 -0.03198431 -0.03198527 -0.03198617 -0.03198701
#> qsec         2.17945242  2.17946705  2.17948017  2.17949194  2.17950249
#> vs          -3.26106048 -3.26120706 -3.26133121 -3.26143510 -3.26152077
#> wt          -7.85641093 -7.85637042 -7.85633297 -7.85629851 -7.85626697
#>                  0.0101      0.0094      0.0088      0.0082      0.0076
#> (Intercept)  8.13071914  8.13005925  8.12943832  8.12885487  8.12830741
#> am           2.17392784  2.17385918  2.17379813  2.17374411  2.17369657
#> carb         2.14397293  2.14394454  2.14391777  2.14389263  2.14386910
#> cyl         -0.55003454 -0.55001364 -0.54999204 -0.54996997 -0.54994762
#> disp         0.02426695  0.02426707  0.02426719  0.02426731  0.02426744
#> drat        -1.91031925 -1.91026175 -1.91020867 -1.91015984 -1.91011504
#> gear         0.73733301  0.73741328  0.73748659  0.73755340  0.73761411
#> hp          -0.03198780 -0.03198853 -0.03198921 -0.03198985 -0.03199044
#> qsec         2.17951196  2.17952044  2.17952806  2.17953490  2.17954105
#> vs          -3.26159006 -3.26164470 -3.26168625 -3.26171616 -3.26173576
#> wt          -7.85623827 -7.85621232 -7.85618903 -7.85616829 -7.85615001
#>                  0.0071     0.0066      0.0062      0.0058
#> (Intercept)  8.12779445  8.1273145  8.12686606  8.12644767
#> am           2.17365499  2.1736189  2.17358779  2.17356129
#> carb         2.14384718  2.1438268  2.14380803  2.14379075
#> cyl         -0.54992518 -0.5499028 -0.54988060 -0.54985870
#> disp         0.02426757  0.0242677  0.02426784  0.02426798
#> drat        -1.91007408 -1.9100368 -1.91000290 -1.90997231
#> gear         0.73766913  0.7377188  0.73776362  0.73780380
#> hp          -0.03199099 -0.0319915 -0.03199197 -0.03199240
#> qsec         2.17954659  2.1795516  2.17955614  2.17956026
#> vs          -3.26174625 -3.2617488 -3.26174427 -3.26173372
#> wt          -7.85613408 -7.8561204 -7.85610885 -7.85609933
#> 
#> $iter
#>   [1]   0   3   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   4   4   4   4   4   4   4   4   4   4   4   4   4   4   4   3
#>  [37]  14  13  13  13  12  12  12   7   1   1   1   1  56   1   1   1   1   1
#>  [55]   1   1   1 288   1   1   1   1   1   1   1   1   1   1   1  80   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.757029246 5.369024743 5.007170445 4.669703915 4.354981500 4.061470323
#>   [7] 3.787740816 3.532459762 3.294383797 3.072353355 2.865287022 2.672176266
#>  [13] 2.492080529 2.324122642 2.167484555 2.021403350 1.885167530 1.758113548
#>  [19] 1.639622579 1.529117505 1.426060103 1.329948424 1.240314352 1.156721316
#>  [25] 1.078762170 1.006057211 0.938252322 0.875017256 0.816044021 0.761045385
#>  [31] 0.709753473 0.661918465 0.617307376 0.575702925 0.536902474 0.500717044
#>  [37] 0.466970392 0.435498150 0.406147032 0.378774082 0.353245976 0.329438380
#>  [43] 0.307235335 0.286528702 0.267217627 0.249208053 0.232412264 0.216748455
#>  [49] 0.202140335 0.188516753 0.175811355 0.163962258 0.152911750 0.142606010
#>  [55] 0.132994842 0.124031435 0.115672132 0.107876217 0.100605721 0.093825232
#>  [61] 0.087501726 0.081604402 0.076104539 0.070975347 0.066191846 0.061730738
#>  [67] 0.057570292 0.053690247 0.050071704 0.046697039 0.043549815 0.040614703
#>  [73] 0.037877408 0.035324598 0.032943838 0.030723534 0.028652870 0.026721763
#>  [79] 0.024920805 0.023241226 0.021674846 0.020214034 0.018851675 0.017581135
#>  [85] 0.016396226 0.015291175 0.014260601 0.013299484 0.012403144 0.011567213
#>  [91] 0.010787622 0.010060572 0.009382523 0.008750173 0.008160440 0.007610454
#>  [97] 0.007097535 0.006619185 0.006173074 0.005757029
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 36
#> 
#> $loss
#>   [1] 836.70952 703.09644 591.30922 498.20697 421.07796 357.58241 305.70279
#>   [8] 263.70061 230.07906 203.55044 183.00800 167.50145 156.21580 148.45293
#>  [15] 143.61563 141.19381 140.69842 140.69842 140.69842 140.69842 139.34717
#>  [22] 130.14307 122.48868 116.15834 110.95755 106.71863 103.29711 100.56867
#>  [29]  98.42633  96.77822  95.54546  94.66042  94.06520  93.71028  93.55340
#>  [36]  93.53761  91.22896  88.95864  87.24829  86.01086  85.17053  84.65384
#>  [43]  84.40740  84.36859  84.36859  84.36859  84.36859  84.36859  77.43028
#>  [50]  77.43027  77.43026  77.43025  77.43025  77.43025  77.43024  77.43024
#>  [57]  77.43024  63.59487  63.59485  63.59484  63.59482  63.59481  63.59480
#>  [64]  63.59479  63.59478  63.59477  63.59476  63.59475  63.59475  62.58063
#>  [71]  62.58062  62.58061  62.58060  62.58060  62.58059  62.58059  62.58058
#>  [78]  62.58058  62.58058  62.58058  62.58057  62.58057  62.58057  62.58057
#>  [85]  62.58057  62.58057  62.58057  62.58057  62.58057  62.58057  62.58057
#>  [92]  62.58057  62.58057  62.58056  62.58056  62.58056  62.58056  62.58056
#>  [99]  62.58056  62.58056
#> 
#> $linear.predictors
#>     5.7570   5.3690   5.0072   4.6697   4.3550   4.0615   3.7877   3.5325
#> 1  19.9381 20.44905 20.92557 21.36998 21.78443 22.17095 22.53142 22.86760
#> 2  19.9381 19.93514 19.93239 19.92982 19.92742 19.92519 19.92311 19.92116
#> 3  19.9381 19.80595 19.68270 19.56777 19.46058 19.36061 19.26739 19.18044
#> 4  19.9381 19.94950 19.96013 19.97005 19.97930 19.98792 19.99597 20.00347
#> 5  19.9381 19.80595 19.68270 19.56777 19.46058 19.36061 19.26739 19.18044
#> 6  19.9381 19.80595 19.68270 19.56777 19.46058 19.36061 19.26739 19.18044
#> 7  19.9381 19.44420 18.98359 18.55403 18.15341 17.77980 17.43136 17.10641
#> 8  19.9381 19.63943 19.36089 19.10112 18.85887 18.63294 18.42223 18.22573
#> 9  19.9381 19.61072 19.30540 19.02067 18.75512 18.50747 18.27652 18.06112
#> 10 19.9381 18.76664 17.67414 16.65527 15.70507 14.81890 13.99247 13.22173
#> 11 19.9381 18.66673 17.48105 16.37528 15.34404 14.38230 13.48537 12.64890
#> 12 19.9381 20.85387 21.70792 22.50441 23.24722 23.93996 24.58602 25.18854
#> 13 19.9381 20.72754 21.46378 22.15040 22.79074 23.38793 23.94487 24.46427
#> 14 19.9381 20.36579 20.76467 21.13666 21.48358 21.80711 22.10885 22.39024
#> 15 19.9381 19.76001 19.59393 19.43904 19.29459 19.15987 19.03424 18.91707
#> 16 19.9381 19.80882 19.68825 19.57581 19.47095 19.37316 19.28196 19.19690
#> 17 19.9381 19.57627 19.23882 18.92412 18.63063 18.35692 18.10166 17.86360
#> 18 19.9381 20.67012 21.35281 21.98949 22.58326 23.13701 23.65344 24.13506
#> 19 19.9381 20.91243 21.82111 22.66854 23.45885 24.19590 24.88328 25.52433
#> 20 19.9381 19.96098 19.98232 20.00223 20.02079 20.03811 20.05425 20.06931
#> 21 19.9381 20.18492 20.41511 20.62979 20.82999 21.01671 21.19084 21.35323
#>      3.2944   3.0724   2.8653    2.6722    2.4921    2.3241    2.1675    2.0214
#> 1  23.18112 23.47351 23.74619 24.000496 24.237661 24.458843 24.665117 24.857490
#> 2  19.91935 19.91766 19.91609 19.914617 19.913246 19.911968 19.910776 19.909664
#> 3  19.09936 19.02373 18.95321 18.887441 18.826103 18.768899 18.715550 18.665797
#> 4  20.01046 20.01699 20.02307 20.028748 20.034040 20.038976 20.043579 20.047871
#> 5  19.09936 19.02373 18.95321 18.887441 18.826103 18.768899 18.715550 18.665797
#> 6  19.09936 19.02373 18.95321 18.887441 18.826103 18.768899 18.715550 18.665797
#> 7  16.80336 16.52074 16.25716 16.011347 15.782101 15.568305 15.368919 15.182970
#> 8  18.04247 17.87156 17.71217 17.563525 17.424896 17.295610 17.175037 17.062591
#> 9  17.86025 17.67291 17.49820 17.335263 17.183308 17.041594 16.909432 16.786176
#> 10 12.50293 11.83258 11.20741 10.624379 10.080638  9.573543  9.100625  8.659580
#> 11 11.86880 11.14128 10.46279  9.830029  9.239913  8.689570  8.176317  7.697656
#> 12 25.75044 26.27448 26.76320 27.218981 27.644044 28.040459 28.410157 28.754939
#> 13 24.94867 25.40042 25.82172 26.214631 26.581059 26.922792 27.241492 27.538714
#> 14 22.65268 22.89742 23.12567 23.338538 23.537058 23.722198 23.894861 24.055887
#> 15 18.80780 18.70589 18.61086 18.522222 18.439563 18.362474 18.290581 18.223533
#> 16 19.11758 19.04360 18.97461 18.910267 18.850262 18.794300 18.742111 18.693439
#> 17 17.64158 17.43453 17.24143 17.061350 16.893403 16.736776 16.590705 16.454478
#> 18 24.58423 25.00312 25.39378 25.758108 26.097885 26.414761 26.710281 26.985884
#> 19 26.12218 26.67973 27.19970 27.684634 28.136883 28.558651 28.951993 29.318825
#> 20 20.08335 20.09645 20.10866 20.120052 20.130675 20.140582 20.149821 20.158437
#> 21 21.50468 21.64592 21.77765 21.900491 22.015057 22.121902 22.221545 22.314473
#>       1.8852    1.7581    1.6396    1.5291    1.4261    1.3299    1.2403
#> 1  24.992330 24.992330 24.992330 24.992330 24.996809 25.029260 25.059524
#> 2  19.908885 19.908885 19.908885 19.908885 19.932457 20.103260 20.262547
#> 3  18.630923 18.630923 18.630923 18.630923 18.616352 18.510769 18.412304
#> 4  20.050880 20.050880 20.050880 20.050880 20.083447 20.319419 20.539483
#> 5  18.630923 18.630923 18.630923 18.630923 18.637263 18.683200 18.726040
#> 6  18.630923 18.630923 18.630923 18.630923 18.647066 18.764027 18.873103
#> 7  15.052632 15.052632 15.052632 15.052632 15.048164 15.015791 14.985600
#> 8  16.983774 16.983774 16.983774 16.983774 16.980471 16.956538 16.934220
#> 9  16.699782 16.699782 16.699782 16.699782 16.703323 16.728981 16.752909
#> 10  8.350436  8.350436  8.350436  8.350436  8.362740  8.451893  8.535037
#> 11  7.362146  7.362146  7.362146  7.362146  7.372912  7.450922  7.523672
#> 12 28.996609 28.996609 28.996609 28.996609 28.995258 28.985468 28.976338
#> 13 27.747047 27.747047 27.747047 27.747047 27.769601 27.933025 28.085432
#> 14 24.168755 24.168755 24.168755 24.168755 24.197003 24.401675 24.592549
#> 15 18.176537 18.176537 18.176537 18.176537 18.160010 18.040255 17.928574
#> 16 18.659323 18.659323 18.659323 18.659323 18.649294 18.576633 18.508869
#> 17 16.358992 16.358992 16.358992 16.358992 16.320592 16.042348 15.782862
#> 18 27.179064 27.179064 27.179064 27.179064 27.185900 27.235430 27.281620
#> 19 29.575951 29.575951 29.575951 29.575951 29.547503 29.341377 29.149146
#> 20 20.164477 20.164477 20.164477 20.164477 20.107066 19.691083 19.303143
#> 21 22.379610 22.379610 22.379610 22.379610 22.386769 22.438647 22.487027
#>       1.1567    1.0788    1.0061    0.9383    0.8750    0.8160    0.7610
#> 1  25.087748 25.114069 25.138617 25.161510 25.182860 25.202772 25.221341
#> 2  20.411099 20.549640 20.678843 20.799338 20.911712 21.016512 21.114250
#> 3  18.320476 18.234836 18.154969 18.080484 18.011019 17.946236 17.885819
#> 4  20.744715 20.936115 21.114615 21.281085 21.436336 21.581123 21.716152
#> 5  18.765992 18.803252 18.838000 18.870407 18.900630 18.928815 18.955101
#> 6  18.974828 19.069697 19.158172 19.240684 19.317635 19.389400 19.456328
#> 7  14.957444 14.931186 14.906697 14.883859 14.862559 14.842696 14.824171
#> 8  16.913405 16.893994 16.875890 16.859007 16.843262 16.828577 16.814883
#> 9  16.775225 16.796036 16.815445 16.833546 16.850427 16.866170 16.880852
#> 10  8.612576  8.684890  8.752330  8.815225  8.873881  8.928583  8.979599
#> 11  7.591520  7.654795  7.713805  7.768838  7.820162  7.868027  7.912666
#> 12 28.967823 28.959882 28.952477 28.945570 28.939129 28.933123 28.927520
#> 13 28.227566 28.360122 28.483743 28.599033 28.706553 28.806826 28.900341
#> 14 24.770558 24.936571 25.091394 25.235783 25.370441 25.496023 25.613142
#> 15 17.824420 17.727286 17.636698 17.552215 17.473426 17.399948 17.331421
#> 16 18.445673 18.386736 18.331771 18.280511 18.232705 18.188121 18.146542
#> 17 15.540865 15.315178 15.104702 14.908410 14.725349 14.554624 14.395407
#> 18 27.324698 27.364872 27.402339 27.437281 27.469868 27.500258 27.528600
#> 19 28.969872 28.802679 28.646756 28.501340 28.365726 28.239251 28.121300
#> 20 18.941349 18.603939 18.289270 17.995808 17.722124 17.466886 17.228849
#> 21 22.532147 22.574226 22.613469 22.650066 22.684198 22.716029 22.745715
#>       0.7098    0.6619    0.6173    0.5757    0.5369    0.5007    0.4670
#> 1  25.238659 25.254810 25.269872 25.283919 25.297019 25.302696 25.322401
#> 2  21.205400 21.290407 21.369685 21.443619 21.512571 21.542453 21.472841
#> 3  17.829473 17.776926 17.727920 17.682216 17.639593 17.621122 17.513795
#> 4  21.842080 21.959522 22.069048 22.171192 22.266453 22.307736 22.429825
#> 5  18.979616 19.002478 19.023800 19.043684 19.062228 19.070265 19.173287
#> 6  19.518745 19.576956 19.631243 19.681872 19.729089 19.749551 19.866431
#> 7  14.806895 14.790783 14.775757 14.761743 14.748675 14.743011 14.695354
#> 8  16.802111 16.790200 16.779092 16.768733 16.759072 16.754885 16.684381
#> 9  16.894545 16.907314 16.919223 16.930330 16.940688 16.945177 16.887950
#> 10  9.027177  9.071548  9.112928  9.151520  9.187511  9.203108  9.264171
#> 11  7.954297  7.993122  8.029330  8.063098  8.094590  8.108237  8.179661
#> 12 28.922296 28.917424 28.912880 28.908642 28.904690 28.902977 28.863651
#> 13 28.987554 29.068888 29.144742 29.215482 29.281455 29.310046 29.320365
#> 14 25.722367 25.824231 25.919229 26.007825 26.090450 26.126257 26.009222
#> 15 17.267513 17.207912 17.152328 17.100491 17.052146 17.031195 16.926867
#> 16 18.107766 18.071602 18.037877 18.006424 17.977091 17.964378 17.863114
#> 17 14.246919 14.108440 13.979293 13.858851 13.746526 13.697847 13.585650
#> 18 27.555032 27.579683 27.602672 27.624112 27.644107 27.652772 27.648073
#> 19 28.011299 27.908711 27.813038 27.723812 27.640600 27.604539 27.700345
#> 20 17.006856 16.799825 16.606746 16.426681 16.258751 16.185975 16.360210
#> 21 22.773400 22.799219 22.823298 22.845754 22.866696 22.875772 22.932407
#>       0.4355    0.4061    0.3788    0.3532    0.3294    0.3072    0.2865
#> 1  25.345399 25.366891 25.386935 25.405549 25.422986 25.439249 25.448693
#> 2  21.391738 21.315985 21.245337 21.179657 21.118199 21.060881 21.027632
#> 3  17.388568 17.271552 17.162420 17.061053 16.966113 16.877567 16.826157
#> 4  22.572333 22.705514 22.829722 22.945064 23.053120 23.153900 23.212426
#> 5  19.293465 19.405755 19.510479 19.607765 19.698870 19.783839 19.833164
#> 6  20.002805 20.130238 20.249084 20.359475 20.462866 20.559294 20.615280
#> 7  14.639756 14.587805 14.539354 14.494347 14.452198 14.412887 14.390065
#> 8  16.602140 16.525298 16.453633 16.387057 16.324712 16.266567 16.232813
#> 9  16.821219 16.758875 16.700732 16.646705 16.596124 16.548950 16.521571
#> 10  9.335432  9.402024  9.464130  9.521810  9.575839  9.626231  9.655491
#> 11  8.263000  8.340876  8.413505  8.480965  8.544149  8.603078  8.637293
#> 12 28.817781 28.774922 28.734951 28.697817 28.663044 28.630614 28.611788
#> 13 29.332480 29.343822 29.354400 29.364187 29.373392 29.381977 29.386980
#> 14 25.872826 25.745418 25.626595 25.516147 25.412780 25.316375 25.260443
#> 15 16.805130 16.691372 16.585280 16.486739 16.394443 16.308363 16.258383
#> 16 17.744975 17.634585 17.531633 17.435999 17.346436 17.262905 17.214410
#> 17 13.454657 13.332230 13.218051 13.112037 13.012705 12.920061 12.866253
#> 18 27.642618 27.637527 27.632779 27.628356 27.624227 27.620376 27.618147
#> 19 27.811982 27.916257 28.013506 28.103910 28.188508 28.267408 28.313181
#> 20 16.563213 16.752825 16.929660 17.094058 17.247889 17.391358 17.474584
#> 21 22.998485 23.060229 23.117814 23.171302 23.221398 23.268121 23.295247
#>       0.2672    0.2492    0.2324    0.2167    0.2021    0.1885    0.1758
#> 1  25.448735 25.448756 25.448766 25.448771 25.201616 25.201568 25.201527
#> 2  21.027564 21.027530 21.027513 21.027505 19.994811 19.994475 19.994188
#> 3  16.825950 16.825846 16.825795 16.825769 17.083801 17.083737 17.083682
#> 4  23.212695 23.212829 23.212895 23.212928 23.659769 23.660129 23.660438
#> 5  19.833348 19.833440 19.833485 19.833508 18.857762 18.857564 18.857395
#> 6  20.615507 20.615620 20.615676 20.615704 20.103516 20.103512 20.103508
#> 7  14.389977 14.389933 14.389911 14.389900 14.666408 14.666438 14.666464
#> 8  16.232689 16.232628 16.232597 16.232582 16.812082 16.812191 16.812285
#> 9  16.521483 16.521439 16.521417 16.521406 17.388111 17.388347 17.388549
#> 10  9.655617  9.655680  9.655711  9.655727  9.865054  9.865218  9.865359
#> 11  8.637433  8.637503  8.637538  8.637555  8.647284  8.647389  8.647478
#> 12 28.611720 28.611686 28.611669 28.611661 28.602835 28.602789 28.602750
#> 13 29.387043 29.387074 29.387089 29.387097 30.348385 30.348774 30.349107
#> 14 25.260305 25.260237 25.260203 25.260186 24.995386 24.995267 24.995166
#> 15 16.258176 16.258073 16.258022 16.257997 16.365204 16.365088 16.364989
#> 16 17.214222 17.214128 17.214081 17.214058 17.690600 17.690627 17.690649
#> 17 12.865989 12.865857 12.865792 12.865760 11.705236 11.704635 11.704122
#> 18 27.618153 27.618156 27.618157 27.618158 27.763180 27.763242 27.763296
#> 19 28.313282 28.313333 28.313358 28.313370 28.133441 28.133394 28.133353
#> 20 17.474757 17.474844 17.474886 17.474908 17.975825 17.975991 17.976134
#> 21 23.295355 23.295409 23.295436 23.295449 22.839694 22.839622 22.839561
#>       0.1640    0.1529    0.1426    0.1330    0.1240    0.1157    0.1079
#> 1  25.201492 25.201462 25.201437 25.201415 25.201396 25.201380 24.597290
#> 2  19.993943 19.993733 19.993553 19.993400 19.993268 19.993156 20.315234
#> 3  17.083635 17.083595 17.083561 17.083531 17.083506 17.083485 17.452349
#> 4  23.660702 23.660928 23.661121 23.661286 23.661427 23.661547 22.643380
#> 5  18.857250 18.857126 18.857021 18.856930 18.856853 18.856787 18.836447
#> 6  20.103504 20.103502 20.103499 20.103497 20.103495 20.103494 20.171636
#> 7  14.666486 14.666505 14.666521 14.666535 14.666546 14.666556 13.522270
#> 8  16.812366 16.812434 16.812493 16.812543 16.812586 16.812623 16.534334
#> 9  17.388722 17.388870 17.388996 17.389104 17.389197 17.389276 17.046960
#> 10  9.865479  9.865581  9.865669  9.865744  9.865808  9.865863 11.508697
#> 11  8.647555  8.647621  8.647677  8.647725  8.647766  8.647801  9.140901
#> 12 28.602717 28.602689 28.602664 28.602644 28.602626 28.602611 30.476462
#> 13 30.349391 30.349635 30.349843 30.350021 30.350173 30.350303 30.629939
#> 14 24.995079 24.995005 24.994941 24.994887 24.994840 24.994801 23.640990
#> 15 16.364904 16.364832 16.364770 16.364716 16.364671 16.364632 17.017352
#> 16 17.690668 17.690685 17.690699 17.690711 17.690721 17.690730 17.686084
#> 17 11.703683 11.703307 11.702986 11.702711 11.702476 11.702276 11.184955
#> 18 27.763342 27.763381 27.763414 27.763443 27.763467 27.763488 28.014483
#> 19 28.133318 28.133288 28.133263 28.133241 28.133223 28.133207 28.852704
#> 20 17.976256 17.976360 17.976449 17.976525 17.976590 17.976645 17.101393
#> 21 22.839509 22.839464 22.839425 22.839393 22.839364 22.839340 22.326140
#>      0.1006   0.0938    0.0875    0.0816   0.0761    0.0710    0.0662    0.0617
#> 1  24.59719 24.59710 24.597013 24.596928 24.59685 24.596770 24.596696 24.596626
#> 2  20.31528 20.31533 20.315375 20.315418 20.31546 20.315498 20.315535 20.315571
#> 3  17.45234 17.45234 17.452332 17.452326 17.45232 17.452316 17.452311 17.452306
#> 4  22.64304 22.64272 22.642405 22.642108 22.64182 22.641552 22.641293 22.641044
#> 5  18.83633 18.83622 18.836120 18.836020 18.83592 18.835832 18.835744 18.835660
#> 6  20.17167 20.17171 20.171741 20.171773 20.17180 20.171833 20.171861 20.171888
#> 7  13.52218 13.52210 13.522022 13.521946 13.52187 13.521806 13.521740 13.521677
#> 8  16.53450 16.53466 16.534811 16.534957 16.53510 16.535230 16.535358 16.535480
#> 9  17.04720 17.04742 17.047640 17.047848 17.04805 17.048237 17.048419 17.048594
#> 10 11.50883 11.50896 11.509077 11.509192 11.50930 11.509408 11.509508 11.509605
#> 11  9.14089  9.14088  9.140869  9.140859  9.14085  9.140841  9.140832  9.140823
#> 12 30.47641 30.47636 30.476308 30.476261 30.47622 30.476172 30.476129 30.476089
#> 13 30.63013 30.63031 30.630484 30.630651 30.63081 30.630964 30.631110 30.631250
#> 14 23.64111 23.64123 23.641344 23.641452 23.64156 23.641656 23.641751 23.641843
#> 15 17.01722 17.01709 17.016974 17.016858 17.01675 17.016642 17.016540 17.016442
#> 16 17.68594 17.68581 17.685682 17.685560 17.68544 17.685330 17.685223 17.685120
#> 17 11.18497 11.18498 11.184995 11.185007 11.18502 11.185030 11.185041 11.185051
#> 18 28.01443 28.01438 28.014337 28.014292 28.01425 28.014208 28.014168 28.014130
#> 19 28.85284 28.85298 28.853104 28.853225 28.85334 28.853453 28.853559 28.853660
#> 20 17.10142 17.10145 17.101480 17.101507 17.10153 17.101559 17.101583 17.101607
#> 21 22.32605 22.32597 22.325887 22.325810 22.32574 22.325665 22.325598 22.325533
#>       0.0576    0.0537   0.0501    0.0467    0.0435    0.0406    0.0379
#> 1  24.596558 24.596493 24.59643 24.563266 24.563122 24.562988 24.562864
#> 2  20.315605 20.315637 20.31567 19.898791 19.898345 19.897943 19.897582
#> 3  17.452302 17.452297 17.45229 17.420762 17.420740 17.420724 17.420714
#> 4  22.640807 22.640580 22.64036 22.957088 22.957628 22.958129 22.958592
#> 5  18.835580 18.835503 18.83543 18.597691 18.597678 18.597670 18.597666
#> 6  20.171914 20.171938 20.17196 19.904947 19.904980 19.905015 19.905050
#> 7  13.521617 13.521560 13.52151 13.393052 13.393100 13.393143 13.393180
#> 8  16.535597 16.535710 16.53582 16.500364 16.500401 16.500430 16.500454
#> 9  17.048761 17.048921 17.04907 16.978992 16.979063 16.979125 16.979179
#> 10 11.509696 11.509784 11.50987 11.760009 11.760025 11.760037 11.760046
#> 11  9.140815  9.140807  9.14080  9.299541  9.299549  9.299554  9.299557
#> 12 30.476050 30.476012 30.47598 30.511579 30.511682 30.511778 30.511866
#> 13 30.631384 30.631512 30.63164 30.476123 30.476197 30.476265 30.476329
#> 14 23.641931 23.642015 23.64210 24.015208 24.015234 24.015251 24.015262
#> 15 17.016349 17.016259 17.01617 16.990970 16.990952 16.990938 16.990928
#> 16 17.685022 17.684928 17.68484 17.509893 17.509979 17.510063 17.510146
#> 17 11.185061 11.185070 11.18508 11.470464 11.470251 11.470046 11.469851
#> 18 28.014094 28.014060 28.01403 27.939200 27.939172 27.939148 27.939126
#> 19 28.853757 28.853850 28.85394 29.256915 29.256883 29.256847 29.256808
#> 20 17.101629 17.101651 17.10167 16.876553 16.876546 16.876545 16.876548
#> 21 22.325471 22.325412 22.32536 22.378592 22.378474 22.378361 22.378253
#>       0.0353    0.0329    0.0307    0.0287    0.0267    0.0249    0.0232
#> 1  24.562749 24.562643 24.562545 24.562454 24.562370 24.562293 24.562221
#> 2  19.897259 19.896971 19.896714 19.896486 19.896285 19.896109 19.895954
#> 3  17.420708 17.420706 17.420707 17.420712 17.420719 17.420729 17.420741
#> 4  22.959021 22.959416 22.959781 22.960117 22.960426 22.960710 22.960970
#> 5  18.597666 18.597670 18.597677 18.597687 18.597698 18.597712 18.597727
#> 6  19.905086 19.905123 19.905159 19.905196 19.905232 19.905267 19.905302
#> 7  13.393213 13.393241 13.393265 13.393286 13.393303 13.393317 13.393329
#> 8  16.500471 16.500484 16.500493 16.500498 16.500499 16.500497 16.500493
#> 9  16.979224 16.979262 16.979295 16.979321 16.979342 16.979359 16.979371
#> 10 11.760052 11.760055 11.760056 11.760055 11.760053 11.760050 11.760046
#> 11  9.299559  9.299559  9.299557  9.299555  9.299551  9.299547  9.299542
#> 12 30.511948 30.512023 30.512093 30.512158 30.512218 30.512273 30.512324
#> 13 30.476388 30.476444 30.476495 30.476542 30.476586 30.476627 30.476664
#> 14 24.015265 24.015263 24.015255 24.015243 24.015227 24.015208 24.015186
#> 15 16.990920 16.990915 16.990913 16.990913 16.990915 16.990918 16.990923
#> 16 17.510227 17.510305 17.510381 17.510455 17.510526 17.510594 17.510660
#> 17 11.469665 11.469487 11.469318 11.469158 11.469006 11.468861 11.468725
#> 18 27.939105 27.939088 27.939071 27.939057 27.939044 27.939033 27.939023
#> 19 29.256767 29.256725 29.256681 29.256637 29.256592 29.256547 29.256503
#> 20 16.876557 16.876569 16.876583 16.876601 16.876621 16.876642 16.876664
#> 21 22.378150 22.378052 22.377959 22.377871 22.377787 22.377707 22.377632
#>       0.0217    0.0202    0.0189    0.0176    0.0164    0.0153    0.0143
#> 1  24.562155 24.562095 24.562039 24.561988 24.561941 24.561898 24.561858
#> 2  19.895820 19.895704 19.895604 19.895520 19.895449 19.895390 19.895343
#> 3  17.420755 17.420770 17.420786 17.420803 17.420821 17.420839 17.420858
#> 4  22.961208 22.961426 22.961624 22.961805 22.961970 22.962119 22.962253
#> 5  18.597743 18.597760 18.597778 18.597796 18.597815 18.597834 18.597853
#> 6  19.905337 19.905370 19.905403 19.905435 19.905466 19.905496 19.905525
#> 7  13.393337 13.393344 13.393348 13.393350 13.393351 13.393350 13.393348
#> 8  16.500487 16.500478 16.500469 16.500457 16.500445 16.500432 16.500418
#> 9  16.979380 16.979386 16.979389 16.979389 16.979387 16.979384 16.979379
#> 10 11.760041 11.760035 11.760030 11.760023 11.760017 11.760010 11.760004
#> 11  9.299537  9.299531  9.299525  9.299518  9.299511  9.299505  9.299498
#> 12 30.512371 30.512415 30.512455 30.512492 30.512527 30.512558 30.512587
#> 13 30.476699 30.476731 30.476761 30.476788 30.476813 30.476836 30.476858
#> 14 24.015161 24.015135 24.015107 24.015078 24.015048 24.015017 24.014987
#> 15 16.990929 16.990936 16.990944 16.990953 16.990962 16.990972 16.990982
#> 16 17.510723 17.510783 17.510841 17.510896 17.510948 17.510998 17.511045
#> 17 11.468595 11.468473 11.468358 11.468250 11.468148 11.468052 11.467962
#> 18 27.939014 27.939007 27.939000 27.938994 27.938989 27.938985 27.938981
#> 19 29.256459 29.256416 29.256374 29.256333 29.256293 29.256254 29.256217
#> 20 16.876688 16.876712 16.876737 16.876762 16.876786 16.876811 16.876836
#> 21 22.377561 22.377493 22.377429 22.377369 22.377312 22.377258 22.377208
#>       0.0133    0.0124    0.0116    0.0108    0.0101    0.0094    0.0088
#> 1  24.561822 24.561789 24.561758 24.561731 24.561705 24.561682 24.561661
#> 2  19.895305 19.895276 19.895255 19.895241 19.895233 19.895231 19.895233
#> 3  17.420877 17.420897 17.420916 17.420935 17.420954 17.420973 17.420992
#> 4  22.962375 22.962484 22.962582 22.962670 22.962748 22.962818 22.962879
#> 5  18.597872 18.597891 18.597909 18.597928 18.597945 18.597963 18.597980
#> 6  19.905552 19.905579 19.905605 19.905629 19.905653 19.905675 19.905696
#> 7  13.393345 13.393340 13.393334 13.393328 13.393321 13.393313 13.393305
#> 8  16.500404 16.500389 16.500374 16.500359 16.500344 16.500329 16.500315
#> 9  16.979372 16.979365 16.979356 16.979347 16.979337 16.979327 16.979316
#> 10 11.759998 11.759992 11.759986 11.759981 11.759975 11.759971 11.759966
#> 11  9.299491  9.299484  9.299477  9.299471  9.299464  9.299458  9.299452
#> 12 30.512614 30.512639 30.512662 30.512683 30.512702 30.512720 30.512736
#> 13 30.476877 30.476895 30.476912 30.476927 30.476941 30.476953 30.476965
#> 14 24.014955 24.014924 24.014894 24.014863 24.014833 24.014804 24.014775
#> 15 16.990993 16.991003 16.991014 16.991025 16.991036 16.991047 16.991057
#> 16 17.511090 17.511132 17.511172 17.511209 17.511244 17.511278 17.511308
#> 17 11.467877 11.467798 11.467724 11.467655 11.467590 11.467530 11.467474
#> 18 27.938978 27.938976 27.938974 27.938972 27.938971 27.938970 27.938969
#> 19 29.256182 29.256148 29.256116 29.256085 29.256056 29.256028 29.256002
#> 20 16.876860 16.876883 16.876906 16.876929 16.876950 16.876971 16.876991
#> 21 22.377160 22.377115 22.377073 22.377033 22.376996 22.376961 22.376928
#>       0.0082    0.0076    0.0071   0.0066    0.0062   0.0058
#> 1  24.561641 24.561624 24.561607 24.56159 24.561579 24.56157
#> 2  19.895239 19.895249 19.895262 19.89528 19.895296 19.89532
#> 3  17.421010 17.421028 17.421045 17.42106 17.421078 17.42109
#> 4  22.962933 22.962980 22.963020 22.96306 22.963085 22.96311
#> 5  18.597996 18.598011 18.598026 18.59804 18.598054 18.59807
#> 6  19.905717 19.905736 19.905754 19.90577 19.905788 19.90580
#> 7  13.393296 13.393287 13.393277 13.39327 13.393257 13.39325
#> 8  16.500300 16.500286 16.500272 16.50026 16.500245 16.50023
#> 9  16.979306 16.979295 16.979284 16.97927 16.979263 16.97925
#> 10 11.759962 11.759959 11.759956 11.75995 11.759951 11.75995
#> 11  9.299446  9.299441  9.299435  9.29943  9.299425  9.29942
#> 12 30.512751 30.512765 30.512778 30.51279 30.512800 30.51281
#> 13 30.476976 30.476986 30.476995 30.47700 30.477010 30.47702
#> 14 24.014747 24.014720 24.014694 24.01467 24.014644 24.01462
#> 15 16.991068 16.991078 16.991088 16.99110 16.991107 16.99112
#> 16 17.511337 17.511364 17.511390 17.51141 17.511435 17.51145
#> 17 11.467421 11.467373 11.467328 11.46729 11.467248 11.46721
#> 18 27.938968 27.938968 27.938968 27.93897 27.938968 27.93897
#> 19 29.255978 29.255955 29.255934 29.25591 29.255896 29.25588
#> 20 16.877010 16.877028 16.877046 16.87706 16.877078 16.87709
#> 21 22.376897 22.376868 22.376841 22.37682 22.376792 22.37677
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 22.8 21.4 18.7 24.4 19.2 17.8 16.4 17.3 15.2 10.4 10.4 30.4 33.9 21.5 15.5
#> [16] 15.2 13.3 27.3 30.4 15.8 21.4
#> 
#> $X
#>               am       carb        cyl       disp       drat       gear
#>  [1,]  1.4142136 -1.2961751 -1.1893017 -0.9948898  0.4687316  0.6495191
#>  [2,] -0.7071068 -1.2961751 -0.1034175  0.2177801 -0.9271045 -0.8660254
#>  [3,] -0.7071068 -0.4181210  0.9824666  1.0423956 -0.8002103 -0.8660254
#>  [4,] -0.7071068 -0.4181210 -1.1893017 -0.6820209  0.1786877  0.6495191
#>  [5,] -0.7071068  1.3379872 -0.1034175 -0.5130556  0.5956258  0.6495191
#>  [6,] -0.7071068  1.3379872 -0.1034175 -0.5130556  0.5956258  0.6495191
#>  [7,] -0.7071068  0.4599331  0.9824666  0.3616836 -0.9452322 -0.8660254
#>  [8,] -0.7071068  0.4599331  0.9824666  0.3616836 -0.9452322 -0.8660254
#>  [9,] -0.7071068  0.4599331  0.9824666  0.3616836 -0.9452322 -0.8660254
#> [10,] -0.7071068  1.3379872  0.9824666  1.9478558 -1.1990206 -0.8660254
#> [11,] -0.7071068  1.3379872  0.9824666  1.8508422 -1.0721264 -0.8660254
#> [12,]  1.4142136 -0.4181210 -1.1893017 -1.2560180  2.4265276  0.6495191
#> [13,]  1.4142136 -1.2961751 -1.1893017 -1.2932066  1.1394580  0.6495191
#> [14,] -0.7071068 -1.2961751 -1.1893017 -0.8970677  0.1968155 -0.8660254
#> [15,] -0.7071068 -0.4181210  0.9824666  0.7028481 -1.5071922 -0.8660254
#> [16,] -0.7071068 -0.4181210  0.9824666  0.5896655 -0.8002103 -0.8660254
#> [17,] -0.7071068  1.3379872  0.9824666  0.9615510  0.2511987 -0.8660254
#> [18,]  1.4142136 -1.2961751 -1.1893017 -1.2293393  0.8856696  0.6495191
#> [19,]  1.4142136 -0.4181210 -1.1893017 -1.0991794  0.3237097  2.1650635
#> [20,]  1.4142136  1.3379872  0.9824666  0.9696354  1.1394580  2.1650635
#> [21,]  1.4142136 -0.4181210 -1.1893017 -0.8897917  0.9400528  0.6495191
#>               hp         qsec         vs           wt
#>  [1,] -0.8020556  0.443318220  0.9534626 -0.877924171
#>  [2,] -0.5168625  1.045581399  0.9534626  0.005073892
#>  [3,]  0.5735816 -0.710414857 -1.0488088  0.227056645
#>  [4,] -1.3221136  1.451927640  0.9534626 -0.019590859
#>  [5,] -0.2987737  0.218376551  0.9534626  0.227056645
#>  [6,] -0.2987737  0.653747524  0.9534626  0.227056645
#>  [7,]  0.6574619 -0.434679907 -1.0488088  0.848608354
#>  [8,]  0.6574619 -0.289556250 -1.0488088  0.513167749
#>  [9,]  0.6574619  0.000691065 -1.0488088  0.562497249
#> [10,]  1.0768635 -0.013821301 -1.0488088  2.012784570
#> [11,]  1.2446241 -0.129920227 -1.0488088  2.184451233
#> [12,] -1.4898742  0.378012574  0.9534626 -1.573470131
#> [13,] -1.2717854  1.379365811  0.9534626 -1.356420328
#> [14,] -0.7349514  1.459183823  0.9534626 -0.734868619
#> [15,]  0.1541800 -0.819257600 -1.0488088  0.305983846
#> [16,]  0.1541800 -0.507241736 -1.0488088  0.222123695
#> [17,]  1.7479060 -1.878660299 -1.0488088  0.621692650
#> [18,] -1.2550093  0.653747524  0.9534626 -1.257761327
#> [19,] -0.4665343 -0.797489051  0.9534626 -1.674102312
#> [20,]  2.0666513 -2.538972941 -1.0488088 -0.039322659
#> [21,] -0.5336386  0.436062037  0.9534626 -0.424092765
#> attr(,"center")
#>  [1]   0.3333333   2.4761905   6.1904762 231.0619048   3.5914286   3.5714286
#>  [7] 140.8095238  17.9990476   0.5238095   3.2098571
#> attr(,"scale")
#>  [1]   0.4714045   1.1388820   1.8418171 123.6940098   0.5516407   0.6598289
#>  [7]  59.6087394   1.3781351   0.4994328   1.0135923
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 32
#> 
#> $lambda.min
#> [1] 0.6619185
#> 
#> $null.dev
#> [1] 39.84331
#> 
#> $Bias
#> [1] 3.812218
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 9.708038