Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 34.836553 31.995564 28.217077 24.462496 21.343270 18.766782 16.653142
#>   [8] 14.933488 13.548505 12.447140 11.585496 10.925860 10.435870 10.087787
#>  [15]  9.857864  9.716133  9.484774  9.222392  9.184154  9.184154  9.209807
#>  [22]  9.267994  9.343699  9.416897  9.392011  9.392628  9.349867  9.200428
#>  [29]  8.877801  8.016348  7.723817  7.569368  7.424387  7.394417  7.289623
#>  [36]  8.410202  8.336673  8.308702  8.392987  8.532126  8.581360  8.581179
#>  [43]  8.634419  9.108555 10.133851 10.313187 10.585871 10.800180 10.963807
#>  [50] 10.767989 10.584233 10.881023 10.370520 15.233704 15.234621 15.391364
#>  [57] 13.303866 13.383683 13.384014 13.668348 13.352336 13.352695 13.281766
#>  [64] 13.413668 13.852185 13.924676 13.924948 14.053956 14.202279 14.445394
#>  [71] 14.618888 14.763851 14.764057 14.764258 14.332510 14.197773 14.340770
#>  [78] 14.340686 14.340602 14.302204 14.323705 14.323246 14.343089 14.342520
#>  [85] 14.341963 14.341419 14.340888 14.340369 14.339864 14.367405 14.366864
#>  [92] 14.366337 14.365824 14.365325 14.365100 14.397565 14.397358 14.417546
#>  [99] 14.418039 14.418500
#> 
#> $cvse
#>   [1] 11.220085 11.072292 10.337293  9.317404  8.422366  7.634753  6.940003
#>   [8]  6.326026  5.782837  5.302244  4.877544  4.503272  4.174953  3.888887
#>  [15]  3.641942  3.431070  3.237741  3.054779  3.015827  3.015827  3.014458
#>  [22]  3.011717  3.005439  2.997557  3.000059  3.001474  2.956416  2.865872
#>  [29]  2.786835  2.689010  2.624666  2.569767  2.480223  2.419899  2.374517
#>  [36]  2.433488  2.402115  2.390176  2.388634  2.388968  2.389933  2.389904
#>  [43]  2.393333  2.459523  2.688638  2.659589  2.676287  2.743519  2.744480
#>  [50]  2.743850  2.672096  2.722613  2.678374  4.121262  4.121506  4.118840
#>  [57]  3.734434  3.725097  3.725087  3.876344  3.892297  3.892342  3.901990
#>  [64]  3.903211  3.917572  3.911734  3.911812  3.905887  4.021823  4.027275
#>  [71]  4.020033  4.027775  4.027818  4.027863  3.809848  3.827723  3.817766
#>  [78]  3.817805  3.817841  3.832065  3.827668  3.827440  3.836708  3.836500
#>  [85]  3.836297  3.836098  3.835904  3.835715  3.835530  3.830557  3.830390
#>  [92]  3.830227  3.830068  3.829913  3.829919  3.849855  3.849953  3.863176
#>  [99]  3.863506  3.863813
#> 
#> $fold
#>  [1]  1  8  2  4  8  9  6  4  9  7  1  3 10  5  5 10  2  1  7  6  3
#> 
#> $lambda
#>   [1] 4.938732190 4.605878170 4.295457397 4.005957945 3.735969787 3.484177927
#>   [7] 3.249355996 3.030360277 2.826124136 2.635652828 2.458018649 2.292356419
#>  [13] 2.137859268 1.993774709 1.859400966 1.734083564 1.617212136 1.508217451
#>  [19] 1.406568642 1.311770623 1.223361674 1.140911191 1.064017596 0.992306371
#>  [25] 0.925428243 0.863057476 0.804890291 0.750643379 0.700052528 0.652871330
#>  [31] 0.608869988 0.567834188 0.529564065 0.493873219 0.460587817 0.429545740
#>  [37] 0.400595794 0.373596979 0.348417793 0.324935600 0.303036028 0.282612414
#>  [43] 0.263565283 0.245801865 0.229235642 0.213785927 0.199377471 0.185940097
#>  [49] 0.173408356 0.161721214 0.150821745 0.140656864 0.131177062 0.122336167
#>  [55] 0.114091119 0.106401760 0.099230637 0.092542824 0.086305748 0.080489029
#>  [61] 0.075064338 0.070005253 0.065287133 0.060886999 0.056783419 0.052956406
#>  [67] 0.049387322 0.046058782 0.042954574 0.040059579 0.037359698 0.034841779
#>  [73] 0.032493560 0.030303603 0.028261241 0.026356528 0.024580186 0.022923564
#>  [79] 0.021378593 0.019937747 0.018594010 0.017340836 0.016172121 0.015082175
#>  [85] 0.014065686 0.013117706 0.012233617 0.011409112 0.010640176 0.009923064
#>  [91] 0.009254282 0.008630575 0.008048903 0.007506434 0.007000525 0.006528713
#>  [97] 0.006088700 0.005678342 0.005295641 0.004938732
#> 
#> $fit
#> $beta
#>             4.9387       4.6059       4.2955      4.0060      3.7360
#> (Intercept)   19.1 20.061756935 20.958694671 21.79518181 22.57529251
#> am             0.0  0.000000000  0.000000000  0.00000000  0.00000000
#> carb           0.0  0.000000000  0.000000000  0.00000000  0.00000000
#> cyl            0.0  0.000000000  0.000000000  0.00000000  0.00000000
#> disp           0.0 -0.003913596 -0.007563429 -0.01096728 -0.01414171
#> drat           0.0  0.000000000  0.000000000  0.00000000  0.00000000
#> gear           0.0  0.000000000  0.000000000  0.00000000  0.00000000
#> hp             0.0  0.000000000  0.000000000  0.00000000  0.00000000
#> qsec           0.0  0.000000000  0.000000000  0.00000000  0.00000000
#> vs             0.0  0.000000000  0.000000000  0.00000000  0.00000000
#> wt             0.0  0.000000000  0.000000000  0.00000000  0.00000000
#>                  3.4842      3.2494      3.0304     2.8261      2.6357
#> (Intercept) 23.30282636 23.98132686 24.61409871 25.2042238 25.75457652
#> am           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> disp        -0.01710221 -0.01986317 -0.02243806 -0.0248394 -0.02707891
#> drat         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#>                  2.4580      2.2924      2.1379      1.9938      1.8594
#> (Intercept) 26.26783727 26.74650596 27.19291399 27.60923561 27.99749854
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.02916747 -0.03111528 -0.03293181 -0.03462591 -0.03620584
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  1.7341      1.6172      1.5082      1.4066      1.3118
#> (Intercept) 28.35959385 28.61339956 28.61339956 28.61339956 28.61339956
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.03767928 -0.03871207 -0.03871207 -0.03871207 -0.03871207
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  1.2234      1.1409      1.0640      0.9923      0.9254
#> (Intercept) 28.61339956 28.61339956 28.61339956 28.61339956 28.74535920
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000 -0.07326332
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.03871207 -0.03871207 -0.03871207 -0.03871207 -0.03832628
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.8631      0.8049     0.7506      0.7001      0.6529
#> (Intercept) 28.87634472 28.99850827 29.1124384 24.78756941 20.63646034
#> am           0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> carb        -0.14598580 -0.21381037 -0.2770638 -0.40495850 -0.52625729
#> cyl          0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> disp        -0.03794333 -0.03758618 -0.0372531 -0.03327277 -0.02945996
#> drat         0.00000000  0.00000000  0.0000000  1.04015497  2.03778556
#> gear         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#>                  0.6089      0.5678     0.5296      0.4939     0.4606
#> (Intercept) 16.76052605 15.57462152 15.5168375 15.46294418 15.4225953
#> am           0.00000000  0.00000000  0.0000000  0.00000000  0.0000000
#> carb        -0.63950732 -0.70725892 -0.7540527 -0.79769315 -0.8303977
#> cyl          0.00000000  0.00000000  0.0000000  0.00000000  0.0000000
#> disp        -0.02589951 -0.02459141 -0.0242417 -0.02391556 -0.0236712
#> drat         2.96924988  3.26778261  3.3002118  3.33045623  3.3531139
#> gear         0.00000000  0.00000000  0.0000000  0.00000000  0.0000000
#> hp           0.00000000  0.00000000  0.0000000  0.00000000  0.0000000
#> qsec         0.00000000  0.00000000  0.0000000  0.00000000  0.0000000
#> vs           0.00000000  0.00000000  0.0000000  0.00000000  0.0000000
#> wt           0.00000000  0.00000000  0.0000000  0.00000000  0.0000000
#>                  0.4295      0.4006      0.3736      0.3484      0.3249
#> (Intercept) 15.42148203 15.42082920 15.42044638 15.42022190 15.42009026
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb        -0.83045621 -0.83049053 -0.83051065 -0.83052244 -0.83052936
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.02366955 -0.02366858 -0.02366801 -0.02366768 -0.02366748
#> drat         3.35336103  3.35350594  3.35359092  3.35364074  3.35366996
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.3030     0.2826      0.2636      0.2458      0.2292
#> (Intercept) 15.42001307 15.4199678 15.41994127 15.41992571 15.41991658
#> am           0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> carb        -0.83053342 -0.8305358 -0.83053719 -0.83053801 -0.83053849
#> cyl          0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> disp        -0.02366737 -0.0236673 -0.02366726 -0.02366724 -0.02366723
#> drat         3.35368710  3.3536971  3.35370303  3.35370649  3.35370852
#> gear         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#>                  0.2138      0.1994      0.1859      0.1734      0.1617
#> (Intercept) 15.41991123 17.60015538 17.60229972 17.60362976 17.60441808
#> am           0.00000000  1.23244239  1.23262190  1.23267827  1.23267044
#> carb        -0.83053877 -0.95105111 -0.95101087 -0.95097358 -0.95094224
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.02366722 -0.02261582 -0.02261814 -0.02261978 -0.02262091
#> drat         3.35370970  2.63282875  2.63233538  2.63203911  2.63187082
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.1508     0.1407     0.1312     0.1223     0.1141     0.1064
#> (Intercept) 17.60485782 -23.369421 -23.374586 -23.378748 -23.382101 -23.384802
#> am           1.23263372   2.850606   2.850984   2.851268   2.851481   2.851642
#> carb        -0.95091750   0.000000   0.000000   0.000000   0.000000   0.000000
#> cyl          0.00000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> disp        -0.02262166   0.000000   0.000000   0.000000   0.000000   0.000000
#> drat         2.63178268   2.560597   2.560800   2.560977   2.561128   2.561256
#> gear         0.00000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> hp           0.00000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> qsec         0.00000000   2.553923   2.554165   2.554358   2.554513   2.554637
#> vs           0.00000000  -4.605520  -4.606221  -4.606789  -4.607250  -4.607625
#> wt           0.00000000  -3.320305  -3.320222  -3.320157  -3.320108  -3.320069
#>                 0.0992      0.0925      0.0863      0.0805      0.0751
#> (Intercept) -23.386980 -26.7091864 -26.7123785 -26.7153300 -42.8737154
#> am            2.851764   2.4848893   2.4856103   2.4861616   3.0182148
#> carb          0.000000   0.0000000   0.0000000   0.0000000   0.0000000
#> cyl           0.000000   0.0000000   0.0000000   0.0000000   0.6392627
#> disp          0.000000   0.0000000   0.0000000   0.0000000   0.0000000
#> drat          2.561363   2.2958133   2.2955757   2.2954398   2.4344761
#> gear          0.000000   0.5569962   0.5569928   0.5570008   1.0335673
#> hp            0.000000   0.0000000   0.0000000   0.0000000   0.0000000
#> qsec          2.554736   2.7044123   2.7046255   2.7048042   3.3320181
#> vs           -4.607928  -4.9015128  -4.9019943  -4.9024102  -5.2177487
#> wt           -3.320039  -3.3658042  -3.3657490  -3.3656921  -3.7490098
#>                  0.0700     0.0653      0.0609      0.0568      0.0530
#> (Intercept) -42.8812039 -42.888234 -42.8948328 -42.9010277 -42.9068430
#> am            3.0185876   3.018938   3.0192662   3.0195746   3.0198641
#> carb          0.0000000   0.000000   0.0000000   0.0000000   0.0000000
#> cyl           0.6395342   0.639789   0.6400283   0.6402529   0.6404637
#> disp          0.0000000   0.000000   0.0000000   0.0000000   0.0000000
#> drat          2.4343256   2.434184   2.4340516   2.4339270   2.4338101
#> gear          1.0338451   1.034106   1.0343509   1.0345807   1.0347965
#> hp            0.0000000   0.000000   0.0000000   0.0000000   0.0000000
#> qsec          3.3323511   3.332664   3.3329571   3.3332325   3.3334911
#> vs           -5.2180029  -5.218242  -5.2184657  -5.2186760  -5.2188735
#> wt           -3.7491987  -3.749376  -3.7495425  -3.7496987  -3.7498454
#>                  0.0494      0.0461      0.0430      0.0401        0.0374
#> (Intercept) -42.9123022 -42.9174269 -42.9222377 -42.9267538 -4.293399e+01
#> am            3.0201359   3.0203910   3.0206305   3.0208554  3.021363e+00
#> carb          0.0000000   0.0000000   0.0000000   0.0000000  0.000000e+00
#> cyl           0.6406616   0.6408474   0.6410218   0.6411855  6.413610e-01
#> disp          0.0000000   0.0000000   0.0000000   0.0000000  0.000000e+00
#> drat          2.4337004   2.4335973   2.4335006   2.4334098  2.433146e+00
#> gear          1.0349991   1.0351892   1.0353677   1.0355353  1.035726e+00
#> hp            0.0000000   0.0000000   0.0000000   0.0000000  2.973862e-06
#> qsec          3.3337338   3.3339617   3.3341756   3.3343764  3.334752e+00
#> vs           -5.2190588  -5.2192329  -5.2193962  -5.2195495 -5.219851e+00
#> wt           -3.7499831  -3.7501124  -3.7502338  -3.7503477 -3.750583e+00
#>                    0.0348        0.0325       0.0303        0.0283       0.0264
#> (Intercept) -48.403817113 -48.411020855 -48.41796561 -48.424660670 -48.43111502
#> am            3.050822028   3.050954334   3.05108183   3.051204698   3.05132312
#> carb          0.000000000   0.000000000   0.00000000   0.000000000   0.00000000
#> cyl           0.614353063   0.614442013   0.61452779   0.614610518   0.61469030
#> disp          0.000000000   0.000000000   0.00000000   0.000000000   0.00000000
#> drat          2.524258709   2.524312510   2.52436439   2.524414413   2.52446263
#> gear          1.057691560   1.057838188   1.05797959   1.058115940   1.05824743
#> hp            0.005759052   0.005762768   0.00576635   0.005769803   0.00577313
#> qsec          3.639701471   3.640057792   3.64040129   3.640732413   3.64105162
#> vs           -5.807811435  -5.808401508  -5.80897030  -5.809518587  -5.81004709
#> wt           -3.996791316  -3.997024047  -3.99724840  -3.997464670  -3.99767316
#>                    0.0246        0.0229        0.0214        0.0199
#> (Intercept) -48.437337292 -48.443335831 -48.449118672 -48.454693561
#> am            3.051437273   3.051547311   3.051653392   3.051755663
#> carb          0.000000000   0.000000000   0.000000000   0.000000000
#> cyl           0.614767239   0.614841441   0.614913003   0.614982016
#> disp          0.000000000   0.000000000   0.000000000   0.000000000
#> drat          2.524509105   2.524553892   2.524597049   2.524638632
#> gear          1.058374215   1.058496474   1.058614362   1.058728034
#> hp            0.005776338   0.005779429   0.005782409   0.005785281
#> qsec          3.641359342   3.641655988   3.641941958   3.642217638
#> vs           -5.810556525  -5.811047578  -5.811520911  -5.811977163
#> wt           -3.997874143  -3.998067898  -3.998254683  -3.998434749
#>                   0.0186        0.0173        0.0162       0.0151       0.0141
#> (Intercept) -48.46006797 -48.465249097 -48.470243888 -48.47505904 -48.47970100
#> am            3.05185427   3.051949337   3.052041004   3.05212939   3.05221462
#> carb          0.00000000   0.000000000   0.000000000   0.00000000   0.00000000
#> cyl           0.61504857   0.615112756   0.615174653   0.61523434   0.61529190
#> disp          0.00000000   0.000000000   0.000000000   0.00000000   0.00000000
#> drat          2.52467870   2.524717293   2.524754477   2.52479030   2.52482481
#> gear          1.05883764   1.058943319   1.059045214   1.05914346   1.05923818
#> hp            0.00578805   0.005790718   0.005793291   0.00579577   0.00579816
#> qsec          3.64248340   3.642739596   3.642986576   3.64322467   3.64345420
#> vs           -5.81241695  -5.812840872  -5.813249497  -5.81364338  -5.81402306
#> wt           -3.99860834  -3.998775689  -3.998937021  -3.99909255  -3.99924249
#>                    0.0131        0.0122        0.0114        0.0106
#> (Intercept) -48.484175996 -48.488490029 -48.492648882 -48.496658133
#> am            3.052296797   3.052376040   3.052452450   3.052526131
#> carb          0.000000000   0.000000000   0.000000000   0.000000000
#> cyl           0.615347407   0.615400930   0.615452539   0.615502303
#> disp          0.000000000   0.000000000   0.000000000   0.000000000
#> drat          2.524858052   2.524890078   2.524920930   2.524950654
#> gear          1.059329503   1.059417551   1.059502438   1.059584278
#> hp            0.005800464   0.005802685   0.005804826   0.005806889
#> qsec          3.643675470   3.643888781   3.644094418   3.644292658
#> vs           -5.814389040  -5.814741824  -5.815081889  -5.815409695
#> wt           -3.999387039  -3.999526391  -3.999660733  -3.999790245
#>                    0.0099        0.0093        0.0086        0.0080
#> (Intercept) -48.500523159 -48.504249145 -48.261984688 -48.257865648
#> am            3.052597177   3.052665685   3.029341509   3.029172228
#> carb          0.000000000   0.000000000  -0.039283513  -0.039381838
#> cyl           0.615550287   0.615596552   0.632948633   0.633049512
#> disp          0.000000000   0.000000000  -0.001694920  -0.001698711
#> drat          2.524979289   2.525006878   2.593985641   2.594215383
#> gear          1.059663178   1.059739243   1.029085459   1.029026974
#> hp            0.005808879   0.005810796   0.006772431   0.006770080
#> qsec          3.644483766   3.644667999   3.612413773   3.612091270
#> vs           -5.815725685  -5.816030287  -5.846941840  -5.846448899
#> wt           -3.999915101  -4.000035468  -3.855111248  -3.854577580
#>                    0.0075        0.0070        0.0065        0.0061
#> (Intercept) -48.253831099 -48.249879533 -48.246009511 -48.242219667
#> am            3.029010971   3.028857405   3.028711200   3.028572036
#> carb         -0.039475620  -0.039565011  -0.039650165  -0.039731229
#> cyl           0.633147387   0.633242174   0.633333805   0.633422232
#> disp         -0.001702318  -0.001705745  -0.001709001  -0.001712091
#> drat          2.594431517   2.594634593   2.594825157   2.595003750
#> gear          1.028971328   1.028918298   1.028867673   1.028819260
#> hp            0.006767722   0.006765363   0.006763009   0.006760662
#> qsec          3.611777506   3.611472330   3.611175587   3.610887121
#> vs           -5.845962074  -5.845481799  -5.845008470  -5.844542445
#> wt           -3.854065126  -3.853573243  -3.853101293  -3.852648640
#>                    0.0057        0.0053        0.0049
#> (Intercept) -48.238508693 -48.234875342 -48.231318419
#> am            3.028439599   3.028313582   3.028193686
#> carb         -0.039808351  -0.039881674  -0.039951339
#> cyl           0.633507418   0.633589345   0.633668003
#> disp         -0.001715023  -0.001717802  -0.001720435
#> drat          2.595170905   2.595327149   2.595472995
#> gear          1.028772874   1.028728345   1.028685518
#> hp            0.006758328   0.006756011   0.006753713
#> qsec          3.610606772   3.610334381   3.610069786
#> vs           -5.844084045  -5.843633557  -5.843191233
#> wt           -3.852214655  -3.851798714  -3.851400205
#> 
#> $iter
#>   [1]   0   5   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1   1   6   5   5   5  39  34  34  12   8   8   7   1
#>  [37]   1   1   1   1   1   1   1   1   1   1  60   1   1   1   1 133   1   1
#>  [55]   1   1   1  35   1   1  85   1   1   1   1   1   1   1   1   1   2  98
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1  91   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.938732190 4.605878170 4.295457397 4.005957945 3.735969787 3.484177927
#>   [7] 3.249355996 3.030360277 2.826124136 2.635652828 2.458018649 2.292356419
#>  [13] 2.137859268 1.993774709 1.859400966 1.734083564 1.617212136 1.508217451
#>  [19] 1.406568642 1.311770623 1.223361674 1.140911191 1.064017596 0.992306371
#>  [25] 0.925428243 0.863057476 0.804890291 0.750643379 0.700052528 0.652871330
#>  [31] 0.608869988 0.567834188 0.529564065 0.493873219 0.460587817 0.429545740
#>  [37] 0.400595794 0.373596979 0.348417793 0.324935600 0.303036028 0.282612414
#>  [43] 0.263565283 0.245801865 0.229235642 0.213785927 0.199377471 0.185940097
#>  [49] 0.173408356 0.161721214 0.150821745 0.140656864 0.131177062 0.122336167
#>  [55] 0.114091119 0.106401760 0.099230637 0.092542824 0.086305748 0.080489029
#>  [61] 0.075064338 0.070005253 0.065287133 0.060886999 0.056783419 0.052956406
#>  [67] 0.049387322 0.046058782 0.042954574 0.040059579 0.037359698 0.034841779
#>  [73] 0.032493560 0.030303603 0.028261241 0.026356528 0.024580186 0.022923564
#>  [79] 0.021378593 0.019937747 0.018594010 0.017340836 0.016172121 0.015082175
#>  [85] 0.014065686 0.013117706 0.012233617 0.011409112 0.010640176 0.009923064
#>  [91] 0.009254282 0.008630575 0.008048903 0.007506434 0.007000525 0.006528713
#>  [97] 0.006088700 0.005678342 0.005295641 0.004938732
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 28
#> 
#> $loss
#>   [1] 671.02000 572.69067 490.42356 421.90721 365.14597 318.41795 280.23839
#>   [8] 249.32791 224.58494 205.06184 189.94412 178.53245 170.22705 164.51415
#>  [15] 160.95426 159.17198 158.80741 158.80741 158.80741 158.80741 158.80741
#>  [22] 158.80741 158.80741 158.80741 153.94535 149.65353 146.13056 143.26234
#>  [29] 123.27328 110.60763 104.48187 103.25409 102.79667 102.56320 102.51041
#>  [36] 102.51040 102.51040 102.51040 102.51040 102.51040 102.51040 102.51040
#>  [43] 102.51040 102.51040 102.51040 102.51040  99.80702  99.80701  99.80701
#>  [50]  99.80701  99.80701  79.50844  79.50844  79.50843  79.50843  79.50843
#>  [57]  79.50843  78.64418  78.64418  78.64417  76.76725  76.76724  76.76723
#>  [64]  76.76722  76.76721  76.76720  76.76719  76.76719  76.76718  76.76718
#>  [71]  76.76683  76.43270  76.43269  76.43268  76.43267  76.43266  76.43265
#>  [78]  76.43264  76.43264  76.43263  76.43262  76.43262  76.43261  76.43261
#>  [85]  76.43260  76.43260  76.43259  76.43259  76.43259  76.43258  76.43258
#>  [92]  76.39048  76.39046  76.39044  76.39042  76.39041  76.39039  76.39038
#>  [99]  76.39037  76.39036
#> 
#> $linear.predictors
#>    4.9387   4.6059   4.2955   4.0060   3.7360   3.4842   3.2494   3.0304
#> 1    19.1 19.43558 19.74855 20.04042 20.31262 20.56647 20.80322 21.02401
#> 2    19.1 19.63909 20.14184 20.61072 21.04799 21.45579 21.83610 22.19079
#> 3    19.1 18.65286 18.23586 17.84696 17.48428 17.14603 16.83059 16.53640
#> 4    19.1 18.65286 18.23586 17.84696 17.48428 17.14603 16.83059 16.53640
#> 5    19.1 19.48763 19.84914 20.18628 20.50070 20.79393 21.06740 21.32244
#> 6    19.1 19.40584 19.69106 19.95707 20.20514 20.43650 20.65226 20.85348
#> 7    19.1 18.98239 18.87270 18.77041 18.67501 18.58604 18.50306 18.42568
#> 8    19.1 18.98239 18.87270 18.77041 18.67501 18.58604 18.50306 18.42568
#> 9    19.1 18.21454 17.38876 16.61863 15.90040 15.23059 14.60591 14.02334
#> 10   19.1 18.26150 17.47952 16.75024 16.07010 15.43581 14.84427 14.29259
#> 11   19.1 18.33977 17.63079 16.96958 16.35294 15.77786 15.24153 14.74135
#> 12   19.1 19.75376 20.36345 20.93206 21.46234 21.95688 22.41810 22.84822
#> 13   19.1 19.76550 20.38614 20.96496 21.50476 22.00819 22.47768 22.91554
#> 14   19.1 19.59173 20.05033 20.47801 20.87687 21.24885 21.59576 21.91929
#> 15   19.1 18.81723 18.55352 18.30759 18.07823 17.86432 17.66484 17.47880
#> 16   19.1 18.87202 18.65941 18.46113 18.27621 18.10376 17.94292 17.79293
#> 17   19.1 19.59095 20.04881 20.47582 20.87404 21.24543 21.59179 21.91480
#> 18   19.1 18.68808 18.30393 17.94567 17.61155 17.29995 17.00935 16.73834
#> 19   19.1 19.49429 19.86200 20.20493 20.52474 20.82301 21.10117 21.36058
#> 20   19.1 18.88376 18.68210 18.49403 18.31864 18.15506 18.00251 17.86024
#> 21   19.1 19.58821 20.04352 20.46814 20.86415 21.23346 21.57788 21.89909
#>      2.8261   2.6357   2.4580   2.2924   2.1379   1.9938   1.8594   1.7341
#> 1  21.22992 21.42195 21.60104 21.76806 21.92382 22.06909 22.20456 22.33091
#> 2  22.52157 22.83005 23.11775 23.38606 23.63628 23.86964 24.08727 24.29023
#> 3  16.26204 16.00617 15.76755 15.54501 15.33746 15.14391 14.96340 14.79505
#> 4  16.26204 16.00617 15.76755 15.54501 15.33746 15.14391 14.96340 14.79505
#> 5  21.56028 21.78210 21.98897 22.18189 22.36182 22.52961 22.68610 22.83204
#> 6  21.04114 21.21615 21.37937 21.53159 21.67354 21.80593 21.92940 22.04455
#> 7  18.35352 18.28621 18.22345 18.16491 18.11032 18.05941 18.01193 17.96765
#> 8  18.35352 18.28621 18.22345 18.16491 18.11032 18.05941 18.01193 17.96765
#> 9  13.48003 12.97333 12.50079 12.06009 11.64910 11.26580 10.90834 10.57497
#> 10 13.77810 13.29828 12.85080 12.43348 12.04428 11.68132 11.34281 11.02712
#> 11 14.27489 13.83986 13.43415 13.05578 12.70292 12.37383 12.06693 11.78071
#> 12 23.24936 23.62347 23.97236 24.29773 24.60118 24.88418 25.14810 25.39423
#> 13 23.32388 23.70470 24.05986 24.39108 24.69998 24.98805 25.25672 25.50727
#> 14 22.22101 22.50240 22.76482 23.00956 23.23780 23.45066 23.64918 23.83431
#> 15 17.30529 17.14348 16.99258 16.85185 16.72060 16.59820 16.48404 16.37758
#> 16 17.65305 17.52259 17.40093 17.28746 17.18164 17.08296 16.99092 16.90509
#> 17 22.21604 22.49698 22.75899 23.00334 23.23122 23.44374 23.64194 23.82678
#> 18 16.48559 16.24988 16.03005 15.82504 15.63385 15.45554 15.28925 15.13417
#> 19 21.60251 21.82814 22.03855 22.23479 22.41780 22.58848 22.74765 22.89610
#> 20 17.72756 17.60383 17.48843 17.38081 17.28044 17.18684 17.09954 17.01813
#> 21 22.19866 22.47803 22.73857 22.98156 23.20816 23.41950 23.61659 23.80040
#>      1.6172   1.5082   1.4066   1.3118   1.2234   1.1409   1.0640   0.9923
#> 1  22.41947 22.41947 22.41947 22.41947 22.41947 22.41947 22.41947 22.41947
#> 2  24.43250 24.43250 24.43250 24.43250 24.43250 24.43250 24.43250 24.43250
#> 3  14.67705 14.67705 14.67705 14.67705 14.67705 14.67705 14.67705 14.67705
#> 4  14.67705 14.67705 14.67705 14.67705 14.67705 14.67705 14.67705 14.67705
#> 5  22.93434 22.93434 22.93434 22.93434 22.93434 22.93434 22.93434 22.93434
#> 6  22.12526 22.12526 22.12526 22.12526 22.12526 22.12526 22.12526 22.12526
#> 7  17.93661 17.93661 17.93661 17.93661 17.93661 17.93661 17.93661 17.93661
#> 8  17.93661 17.93661 17.93661 17.93661 17.93661 17.93661 17.93661 17.93661
#> 9  10.34130 10.34130 10.34130 10.34130 10.34130 10.34130 10.34130 10.34130
#> 10 10.80585 10.80585 10.80585 10.80585 10.80585 10.80585 10.80585 10.80585
#> 11 11.58009 11.58009 11.58009 11.58009 11.58009 11.58009 11.58009 11.58009
#> 12 25.56676 25.56676 25.56676 25.56676 25.56676 25.56676 25.56676 25.56676
#> 13 25.68290 25.68290 25.68290 25.68290 25.68290 25.68290 25.68290 25.68290
#> 14 23.96408 23.96408 23.96408 23.96408 23.96408 23.96408 23.96408 23.96408
#> 15 16.30296 16.30296 16.30296 16.30296 16.30296 16.30296 16.30296 16.30296
#> 16 16.84493 16.84493 16.84493 16.84493 16.84493 16.84493 16.84493 16.84493
#> 17 23.95634 23.95634 23.95634 23.95634 23.95634 23.95634 23.95634 23.95634
#> 18 15.02546 15.02546 15.02546 15.02546 15.02546 15.02546 15.02546 15.02546
#> 19 23.00015 23.00015 23.00015 23.00015 23.00015 23.00015 23.00015 23.00015
#> 20 16.96107 16.96107 16.96107 16.96107 16.96107 16.96107 16.96107 16.96107
#> 21 23.92924 23.92924 23.92924 23.92924 23.92924 23.92924 23.92924 23.92924
#>      0.9254   0.8631   0.8049   0.7506   0.7001   0.6529   0.6089   0.5678
#> 1  22.32010 22.22147 22.12948 22.04369 21.90070 21.76520 21.63865 21.55531
#> 2  24.53286 24.63248 24.72539 24.81204 24.79375 24.77400 24.75548 24.79245
#> 3  14.80137 14.92477 15.03986 15.14720 15.27594 15.39738 15.51083 15.60071
#> 4  14.65485 14.63280 14.61224 14.59307 14.52844 14.46714 14.40997 14.38226
#> 5  22.97637 23.01809 23.05699 23.09328 22.93471 22.78160 22.63859 22.61066
#> 6  22.02882 21.93310 21.84382 21.76056 21.66863 21.58206 21.50120 21.43377
#> 7  17.95518 17.97362 17.99081 18.00684 17.58934 17.18863 16.81452 16.70263
#> 8  17.95518 17.97362 17.99081 18.00684 17.58934 17.18863 16.81452 16.70263
#> 9  10.36230 10.38315 10.40259 10.42072 10.51064 10.59704 10.67783 10.71304
#> 10 10.82222 10.83847 10.85362 10.86776 10.98273 11.09321 11.19647 11.23689
#> 11 11.58874 11.59733 11.60535 11.61282 11.88742 12.15110 12.39739 12.48030
#> 12 25.65582 25.74422 25.82667 25.90356 26.00788 26.10587 26.19727 26.26457
#> 13 25.69753 25.71206 25.72561 25.73825 26.58687 27.40011 28.15932 28.40870
#> 14 24.06911 24.17336 24.27060 24.36128 24.23513 24.11187 23.99671 24.00473
#> 15 16.41108 16.51839 16.61848 16.71183 16.26774 15.83997 15.44060 15.35912
#> 16 16.94764 17.04960 17.14469 17.23337 17.13922 17.04714 16.96120 16.97783
#> 17 23.98818 24.01979 24.04927 24.07676 24.58283 25.06730 25.51958 25.67803
#> 18 14.99978 14.97429 14.95052 14.92835 15.87845 16.79044 17.64200 17.90404
#> 19 22.74847 22.49865 22.26565 22.04836 21.29863 20.58401 19.91674 19.59469
#> 20 16.62304 16.28751 15.97459 15.68275 15.21495 14.77271 14.35986 14.08249
#> 21 23.96135 23.99323 24.02296 24.05069 24.22668 24.39459 24.55129 24.61513
#>      0.5296   0.4939   0.4606   0.4295   0.4006   0.3736   0.3484   0.3249
#> 1  21.49278 21.43446 21.39076 21.39064 21.39057 21.39053 21.39050 21.39049
#> 2  24.85050 24.90463 24.94520 24.94515 24.94513 24.94512 24.94511 24.94510
#> 3  15.67739 15.74889 15.80248 15.80262 15.80270 15.80275 15.80278 15.80280
#> 4  14.36729 14.35333 14.34287 14.34291 14.34293 14.34295 14.34295 14.34296
#> 5  22.63026 22.64853 22.66223 22.66215 22.66210 22.66208 22.66206 22.66205
#> 6  21.37455 21.31931 21.27792 21.27782 21.27776 21.27772 21.27770 21.27769
#> 7  16.70047 16.69845 16.69695 16.69687 16.69683 16.69680 16.69679 16.69678
#> 8  16.70047 16.69845 16.69695 16.69687 16.69683 16.69680 16.69679 16.69678
#> 9  10.72816 10.74226 10.75282 10.75298 10.75307 10.75312 10.75315 10.75317
#> 10 11.25008 11.26238 11.27160 11.27175 11.27184 11.27189 11.27192 11.27194
#> 11 12.49396 12.50670 12.51624 12.51641 12.51652 12.51658 12.51661 12.51663
#> 12 26.31983 26.37136 26.40998 26.40995 26.40993 26.40991 26.40991 26.40990
#> 13 28.44368 28.47630 28.50074 28.50085 28.50092 28.50096 28.50098 28.50100
#> 14 24.06214 24.11568 24.15581 24.15575 24.15571 24.15569 24.15568 24.15567
#> 15 15.40845 15.45447 15.48895 15.48893 15.48892 15.48891 15.48890 15.48890
#> 16 17.03492 17.08816 17.12807 17.12811 17.12814 17.12816 17.12817 17.12818
#> 17 25.71239 25.74444 25.76845 25.76851 25.76855 25.76857 25.76858 25.76859
#> 18 17.91868 17.93233 17.94256 17.94283 17.94299 17.94309 17.94314 17.94317
#> 19 19.42424 19.26528 19.14616 19.14583 19.14563 19.14552 19.14545 19.14541
#> 20 13.87041 13.67263 13.52441 13.52420 13.52407 13.52400 13.52396 13.52393
#> 21 24.63936 24.66195 24.67888 24.67887 24.67886 24.67885 24.67885 24.67885
#>      0.3030   0.2826   0.2636   0.2458   0.2292   0.2138   0.1994   0.1859
#> 1  21.39048 21.39047 21.39047 21.39047 21.39047 21.39047 21.67789 21.67808
#> 2  24.94510 24.94510 24.94510 24.94510 24.94510 24.94510 25.57543 25.57564
#> 3  15.80281 15.80281 15.80282 15.80282 15.80282 15.80282 15.84977 15.84960
#> 4  14.34296 14.34296 14.34296 14.34296 14.34296 14.34297 14.10564 14.10552
#> 5  22.66205 22.66205 22.66204 22.66204 22.66204 22.66204 22.09545 22.09551
#> 6  21.27768 21.27768 21.27767 21.27767 21.27767 21.27767 20.32623 20.32621
#> 7  16.69677 16.69677 16.69677 16.69677 16.69676 16.69676 16.59234 16.59245
#> 8  16.69677 16.69677 16.69677 16.69677 16.69676 16.69676 16.59234 16.59245
#> 9  10.75318 10.75319 10.75319 10.75320 10.75320 10.75320 10.83547 10.83524
#> 10 11.27195 11.27196 11.27196 11.27196 11.27196 11.27196 11.29116 11.29092
#> 11 12.51665 12.51665 12.51666 12.51666 12.51666 12.51666 12.34903 12.34872
#> 12 26.40990 26.40990 26.40990 26.40990 26.40990 26.40990 26.84362 26.84379
#> 13 28.50100 28.50101 28.50101 28.50101 28.50101 28.50101 28.19832 28.19812
#> 14 24.15567 24.15567 24.15567 24.15567 24.15567 24.15567 23.67441 23.67449
#> 15 15.48890 15.48890 15.48890 15.48890 15.48890 15.48890 15.77283 15.77295
#> 16 17.12818 17.12818 17.12818 17.12818 17.12818 17.12818 17.11625 17.11622
#> 17 25.76860 25.76860 25.76860 25.76860 25.76860 25.76860 25.87324 25.87318
#> 18 17.94319 17.94320 17.94321 17.94321 17.94322 17.94322 18.20078 18.20037
#> 19 19.14539 19.14538 19.14537 19.14537 19.14536 19.14536 19.37784 19.37828
#> 20 13.52392 13.52391 13.52391 13.52390 13.52390 13.52390 13.73704 13.73724
#> 21 24.67885 24.67885 24.67885 24.67885 24.67885 24.67885 25.01491 25.01500
#>      0.1734   0.1617   0.1508   0.1407   0.1312   0.1223   0.1141   0.1064
#> 1  21.67820 21.67827 21.67831 22.80588 22.80609 22.80625 22.80638 22.80648
#> 2  25.57575 25.57579 25.57580 24.55936 24.55935 24.55933 24.55931 24.55928
#> 3  15.84948 15.84940 15.84934 16.74238 16.74226 16.74216 16.74209 16.74203
#> 4  14.10546 14.10543 14.10541 13.45074 13.45036 13.45006 13.44982 13.44963
#> 5  22.09558 22.09565 22.09570 21.96035 21.96033 21.96033 21.96032 21.96032
#> 6  20.32625 20.32632 20.32639 18.90989 18.90968 18.90952 18.90939 18.90929
#> 7  16.59253 16.59259 16.59263 15.41623 15.41623 15.41624 15.41625 15.41625
#> 8  16.59253 16.59259 16.59263 17.05592 17.05594 17.05597 17.05599 17.05600
#> 9  10.83507 10.83496 10.83489 12.62106 12.62128 12.62145 12.62158 12.62168
#> 10 11.29075 11.29064 11.29057 11.81394 11.81415 11.81431 11.81444 11.81454
#> 11 12.34852 12.34839 12.34832 11.64361 11.64376 11.64388 11.64398 11.64406
#> 12 26.84388 26.84391 26.84392 27.74311 27.74334 27.74352 27.74366 27.74377
#> 13 28.19800 28.19792 28.19789 29.43577 29.43589 29.43600 29.43609 29.43617
#> 14 23.67456 23.67463 23.67468 24.41871 24.41864 24.41859 24.41856 24.41853
#> 15 15.77302 15.77305 15.77306 15.09503 15.09480 15.09461 15.09446 15.09434
#> 16 17.11619 17.11617 17.11615 17.47408 17.47402 17.47398 17.47395 17.47393
#> 17 25.87313 25.87310 25.87307 26.36969 26.37002 26.37029 26.37051 26.37069
#> 18 18.20007 18.19988 18.19974 16.79342 16.79326 16.79313 16.79303 16.79295
#> 19 19.37858 19.37878 19.37890 19.13911 19.13903 19.13897 19.13891 19.13887
#> 20 13.73738 13.73748 13.73754 13.97948 13.97924 13.97904 13.97888 13.97874
#> 21 25.01505 25.01506 25.01506 23.67224 23.67231 23.67237 23.67240 23.67243
#>      0.0992   0.0925   0.0863   0.0805   0.0751   0.0700   0.0653   0.0609
#> 1  22.80656 22.65358 22.65382 22.65401 22.63141 22.63144 22.63146 22.63148
#> 2  24.55926 24.46150 24.46172 24.46187 24.30196 24.30196 24.30197 24.30198
#> 3  16.74199 16.64435 16.64421 16.64409 16.82404 16.82410 16.82416 16.82421
#> 4  13.44948 13.15333 13.15294 13.15261 12.55096 12.55059 12.55025 12.54993
#> 5  21.96032 22.44017 22.44004 22.43996 22.26409 22.26405 22.26401 22.26397
#> 6  18.90922 19.15190 19.15150 19.15120 19.50008 19.50013 19.50017 19.50022
#> 7  15.41626 15.36790 15.36790 15.36790 15.53357 15.53365 15.53373 15.53380
#> 8  17.05602 17.05316 17.05318 17.05319 17.47464 17.47485 17.47505 17.47524
#> 9  12.62177 12.64340 12.64362 12.64381 12.70149 12.70156 12.70162 12.70169
#> 10 11.81462 11.78575 11.78593 11.78609 11.68645 11.68642 11.68640 11.68638
#> 11 11.64412 11.49792 11.49796 11.49801 11.20974 11.20957 11.20940 11.20924
#> 12 27.74385 27.71923 27.71957 27.71983 28.17730 28.17758 28.17785 28.17810
#> 13 29.43623 29.07048 29.07038 29.07032 29.27436 29.27431 29.27426 29.27421
#> 14 24.41851 24.37338 24.37322 24.37309 24.00622 24.00604 24.00587 24.00571
#> 15 15.09425 15.07405 15.07398 15.07390 15.07487 15.07493 15.07498 15.07502
#> 16 17.47392 17.41841 17.41834 17.41827 17.77575 17.77591 17.77605 17.77619
#> 17 26.37084 26.69200 26.69214 26.69228 26.27594 26.27579 26.27564 26.27551
#> 18 16.79289 16.79340 16.79317 16.79301 17.12983 17.12987 17.12990 17.12994
#> 19 19.13882 19.46664 19.46675 19.46683 19.22224 19.22223 19.22223 19.22223
#> 20 13.97862 14.15636 14.15634 14.15631 14.30798 14.30808 14.30818 14.30826
#> 21 23.67245 23.48310 23.48328 23.48342 23.17706 23.17693 23.17682 23.17671
#>      0.0568   0.0530   0.0494   0.0461   0.0430   0.0401   0.0374   0.0348
#> 1  22.63150 22.63152 22.63153 22.63155 22.63157 22.63158 22.63152 22.47989
#> 2  24.30198 24.30199 24.30199 24.30199 24.30200 24.30200 24.30213 24.24365
#> 3  16.82426 16.82431 16.82435 16.82439 16.82443 16.82447 16.82447 16.84209
#> 4  12.54962 12.54934 12.54907 12.54882 12.54858 12.54836 12.54809 12.58225
#> 5  22.26393 22.26390 22.26387 22.26384 22.26381 22.26378 22.26367 22.19239
#> 6  19.50026 19.50030 19.50034 19.50037 19.50041 19.50044 19.50033 19.35011
#> 7  15.53387 15.53393 15.53399 15.53404 15.53409 15.53414 15.53417 15.53405
#> 8  17.47541 17.47557 17.47573 17.47587 17.47601 17.47613 17.47632 17.62090
#> 9  12.70175 12.70180 12.70185 12.70190 12.70195 12.70199 12.70207 12.71944
#> 10 11.68636 11.68634 11.68632 11.68630 11.68628 11.68627 11.68626 11.67594
#> 11 11.20909 11.20895 11.20882 11.20870 11.20859 11.20848 11.20833 11.20277
#> 12 28.17833 28.17855 28.17875 28.17894 28.17912 28.17929 28.17963 28.27849
#> 13 29.27417 29.27413 29.27409 29.27405 29.27402 29.27399 29.27384 29.22389
#> 14 24.00556 24.00541 24.00528 24.00515 24.00504 24.00493 24.00490 24.29558
#> 15 15.07507 15.07511 15.07515 15.07519 15.07522 15.07525 15.07521 14.84795
#> 16 17.77632 17.77644 17.77655 17.77665 17.77675 17.77685 17.77688 17.73721
#> 17 26.27539 26.27527 26.27516 26.27506 26.27496 26.27487 26.27466 26.32930
#> 18 17.12997 17.13000 17.13003 17.13006 17.13008 17.13011 17.13010 17.12889
#> 19 19.22222 19.22222 19.22222 19.22221 19.22221 19.22221 19.22221 19.11149
#> 20 14.30835 14.30842 14.30850 14.30856 14.30863 14.30869 14.30901 14.58654
#> 21 23.17661 23.17651 23.17642 23.17634 23.17626 23.17618 23.17618 23.11718
#>      0.0325   0.0303   0.0283   0.0264   0.0246   0.0229   0.0214   0.0199
#> 1  22.47981 22.47974 22.47966 22.47959 22.47953 22.47946 22.47940 22.47934
#> 2  24.24357 24.24350 24.24343 24.24336 24.24330 24.24324 24.24317 24.24312
#> 3  16.84212 16.84215 16.84218 16.84221 16.84223 16.84226 16.84229 16.84231
#> 4  12.58209 12.58194 12.58180 12.58166 12.58152 12.58139 12.58127 12.58115
#> 5  22.19235 22.19232 22.19228 22.19225 22.19221 22.19218 22.19215 22.19212
#> 6  19.35004 19.34997 19.34990 19.34984 19.34978 19.34972 19.34966 19.34960
#> 7  15.53409 15.53412 15.53415 15.53418 15.53421 15.53424 15.53427 15.53430
#> 8  17.62108 17.62126 17.62144 17.62160 17.62176 17.62191 17.62206 17.62221
#> 9  12.71950 12.71955 12.71959 12.71964 12.71969 12.71973 12.71977 12.71981
#> 10 11.67593 11.67593 11.67593 11.67592 11.67592 11.67591 11.67591 11.67591
#> 11 11.20271 11.20265 11.20259 11.20254 11.20249 11.20244 11.20239 11.20234
#> 12 28.27866 28.27883 28.27899 28.27914 28.27929 28.27943 28.27957 28.27970
#> 13 29.22385 29.22382 29.22378 29.22375 29.22372 29.22369 29.22366 29.22363
#> 14 24.29570 24.29581 24.29592 24.29603 24.29613 24.29622 24.29632 24.29641
#> 15 14.84780 14.84765 14.84751 14.84737 14.84723 14.84711 14.84698 14.84686
#> 16 17.73725 17.73729 17.73733 17.73736 17.73740 17.73743 17.73746 17.73749
#> 17 26.32934 26.32939 26.32943 26.32947 26.32951 26.32955 26.32959 26.32962
#> 18 17.12890 17.12891 17.12892 17.12893 17.12894 17.12895 17.12896 17.12897
#> 19 19.11141 19.11133 19.11126 19.11119 19.11111 19.11105 19.11098 19.11092
#> 20 14.58672 14.58690 14.58706 14.58722 14.58738 14.58753 14.58767 14.58781
#> 21 23.11707 23.11696 23.11685 23.11675 23.11666 23.11656 23.11647 23.11639
#>      0.0186   0.0173   0.0162   0.0151   0.0141   0.0131   0.0122   0.0114
#> 1  22.47928 22.47923 22.47917 22.47912 22.47907 22.47902 22.47898 22.47893
#> 2  24.24306 24.24301 24.24295 24.24290 24.24286 24.24281 24.24276 24.24272
#> 3  16.84233 16.84236 16.84238 16.84240 16.84242 16.84244 16.84246 16.84248
#> 4  12.58103 12.58092 12.58081 12.58070 12.58060 12.58051 12.58041 12.58032
#> 5  22.19209 22.19206 22.19204 22.19201 22.19199 22.19196 22.19194 22.19192
#> 6  19.34955 19.34950 19.34945 19.34940 19.34936 19.34931 19.34927 19.34923
#> 7  15.53432 15.53435 15.53437 15.53439 15.53442 15.53444 15.53446 15.53448
#> 8  17.62234 17.62248 17.62261 17.62273 17.62285 17.62296 17.62307 17.62318
#> 9  12.71985 12.71989 12.71992 12.71996 12.71999 12.72002 12.72006 12.72009
#> 10 11.67590 11.67590 11.67590 11.67589 11.67589 11.67589 11.67588 11.67588
#> 11 11.20230 11.20225 11.20221 11.20217 11.20213 11.20209 11.20206 11.20202
#> 12 28.27983 28.27995 28.28007 28.28018 28.28029 28.28040 28.28050 28.28060
#> 13 29.22360 29.22357 29.22355 29.22352 29.22350 29.22347 29.22345 29.22343
#> 14 24.29650 24.29658 24.29666 24.29674 24.29682 24.29689 24.29696 24.29703
#> 15 14.84675 14.84664 14.84653 14.84643 14.84633 14.84623 14.84614 14.84605
#> 16 17.73752 17.73755 17.73758 17.73760 17.73763 17.73766 17.73768 17.73770
#> 17 26.32966 26.32969 26.32972 26.32976 26.32978 26.32981 26.32984 26.32987
#> 18 17.12897 17.12898 17.12899 17.12900 17.12900 17.12901 17.12901 17.12902
#> 19 19.11086 19.11080 19.11074 19.11069 19.11064 19.11059 19.11054 19.11049
#> 20 14.58795 14.58808 14.58820 14.58832 14.58844 14.58855 14.58866 14.58876
#> 21 23.11630 23.11622 23.11614 23.11607 23.11599 23.11592 23.11586 23.11579
#>      0.0106   0.0099   0.0093   0.0086   0.0080   0.0075   0.0070   0.0065
#> 1  22.47889 22.47885 22.47881 22.47452 22.47457 22.47462 22.47466 22.47471
#> 2  24.24268 24.24264 24.24260 24.24606 24.24607 24.24607 24.24608 24.24610
#> 3  16.84250 16.84251 16.84253 16.77805 16.77790 16.77776 16.77762 16.77749
#> 4  12.58024 12.58015 12.58007 12.56538 12.56533 12.56529 12.56524 12.56520
#> 5  22.19190 22.19188 22.19186 22.15417 22.15415 22.15413 22.15412 22.15411
#> 6  19.34919 19.34915 19.34911 19.37837 19.37868 19.37898 19.37926 19.37954
#> 7  15.53450 15.53452 15.53453 15.65182 15.65207 15.65231 15.65254 15.65276
#> 8  17.62328 17.62338 17.62348 17.68504 17.68505 17.68505 17.68505 17.68505
#> 9  12.72011 12.72014 12.72017 12.63232 12.63208 12.63185 12.63163 12.63142
#> 10 11.67588 11.67588 11.67587 11.65318 11.65313 11.65307 11.65302 11.65297
#> 11 11.20199 11.20195 11.20192 11.24487 11.24500 11.24512 11.24523 11.24533
#> 12 28.28070 28.28079 28.28088 28.27877 28.27866 28.27856 28.27846 28.27836
#> 13 29.22341 29.22339 29.22337 29.17809 29.17812 29.17815 29.17818 29.17821
#> 14 24.29709 24.29716 24.29722 24.30350 24.30327 24.30305 24.30283 24.30261
#> 15 14.84597 14.84588 14.84580 14.81800 14.81807 14.81814 14.81821 14.81829
#> 16 17.73772 17.73775 17.73777 17.73441 17.73443 17.73446 17.73449 17.73451
#> 17 26.32989 26.32992 26.32994 26.34713 26.34710 26.34707 26.34703 26.34699
#> 18 17.12903 17.12903 17.12904 17.11816 17.11826 17.11837 17.11847 17.11856
#> 19 19.11045 19.11040 19.11036 19.11817 19.11819 19.11822 19.11824 19.11826
#> 20 14.58886 14.58896 14.58905 14.58190 14.58166 14.58143 14.58120 14.58098
#> 21 23.11573 23.11567 23.11561 23.15806 23.15819 23.15832 23.15844 23.15855
#>      0.0061   0.0057   0.0053   0.0049
#> 1  22.47475 22.47480 22.47484 22.47488
#> 2  24.24611 24.24612 24.24614 24.24616
#> 3  16.77737 16.77725 16.77714 16.77703
#> 4  12.56516 12.56512 12.56509 12.56505
#> 5  22.15410 22.15409 22.15408 22.15407
#> 6  19.37981 19.38007 19.38031 19.38055
#> 7  15.65296 15.65316 15.65334 15.65352
#> 8  17.68504 17.68503 17.68502 17.68501
#> 9  12.63123 12.63104 12.63086 12.63068
#> 10 11.65293 11.65288 11.65284 11.65279
#> 11 11.24543 11.24553 11.24561 11.24570
#> 12 28.27827 28.27818 28.27809 28.27800
#> 13 29.17823 29.17826 29.17829 29.17831
#> 14 24.30240 24.30220 24.30199 24.30180
#> 15 14.81836 14.81844 14.81851 14.81859
#> 16 17.73454 17.73456 17.73459 17.73461
#> 17 26.34694 26.34690 26.34685 26.34681
#> 18 17.11866 17.11875 17.11883 17.11892
#> 19 19.11828 19.11831 19.11833 19.11835
#> 20 14.58077 14.58056 14.58037 14.58017
#> 21 23.15867 23.15877 23.15888 23.15898
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 22.8 18.7 14.3 24.4 17.8 16.4 17.3 10.4 10.4 14.7 32.4 30.4 21.5 15.5
#> [16] 15.2 26.0 15.8 19.7 15.0 21.4
#> 
#> $X
#>               am        carb        cyl       disp        drat       gear
#>  [1,]  1.1547005  0.53611602 -0.2102353 -0.6721296  0.54773673  0.3735437
#>  [2,]  1.1547005 -1.24153183 -1.3139708 -1.0797297  0.45702323  0.3735437
#>  [3,] -0.8660254 -0.64898255  0.8935002  0.8955631 -0.81296571 -0.9338592
#>  [4,] -0.8660254  0.53611602  0.8935002  0.8955631 -0.70410952 -0.9338592
#>  [5,] -0.8660254 -0.64898255 -1.3139708 -0.7763811  0.16674005  0.3735437
#>  [6,] -0.8660254  0.53611602 -0.2102353 -0.6125573  0.58402213  0.3735437
#>  [7,] -0.8660254 -0.05643326  0.8935002  0.2355645 -0.95810731 -0.9338592
#>  [8,] -0.8660254 -0.05643326  0.8935002  0.2355645 -0.95810731 -0.9338592
#>  [9,] -0.8660254  0.53611602  0.8935002  1.7734710 -1.21210510 -0.9338592
#> [10,] -0.8660254  0.53611602  0.8935002  1.6794095 -1.08510620 -0.9338592
#> [11,] -0.8660254  0.53611602  0.8935002  1.5226402 -0.66782412 -0.9338592
#> [12,]  1.1547005 -1.24153183 -1.3139708 -1.3093967  0.87430532  0.3735437
#> [13,]  1.1547005 -0.64898255 -1.3139708 -1.3329121  2.41643475  0.3735437
#> [14,] -0.8660254 -1.24153183 -1.3139708 -0.9848843  0.18488274 -0.9338592
#> [15,] -0.8660254 -0.64898255  0.8935002  0.5663477 -1.52053098 -0.9338592
#> [16,] -0.8660254 -0.64898255  0.8935002  0.4566092 -0.81296571 -0.9338592
#> [17,]  1.1547005 -0.64898255 -1.3139708 -0.9833166  1.50929979  1.6809466
#> [18,]  1.1547005  0.53611602  0.8935002  0.8250169  1.12830311  1.6809466
#> [19,]  1.1547005  1.72121458 -0.2102353 -0.7897065  0.03974115  1.6809466
#> [20,]  1.1547005  2.90631314  0.8935002  0.4330938 -0.10540044  1.6809466
#> [21,]  1.1547005 -0.64898255 -1.3139708 -0.9778297  0.92873341  0.3735437
#>               hp        qsec         vs          wt
#>  [1,] -0.6567986 -0.67049838 -0.7071068 -0.70648085
#>  [2,] -0.8919813  0.74017272  1.4142136 -1.00233497
#>  [3,]  0.2424292 -0.30306777 -0.7071068  0.10218707
#>  [4,]  1.2108284 -1.07729655 -0.7071068  0.23039052
#>  [5,] -1.3208438  1.65218799  1.4142136 -0.14435803
#>  [6,] -0.4769531  0.93044929  1.4142136  0.10218707
#>  [7,]  0.3116006 -0.05373985 -0.7071068  0.72348072
#>  [8,]  0.3116006  0.07748537 -0.7071068  0.38817939
#>  [9,]  0.6574574  0.32681328 -0.7071068  1.88717358
#> [10,]  0.7958002  0.22183311 -0.7071068  2.05876897
#> [11,]  1.0033143 -0.04061733 -0.7071068  1.98086072
#> [12,] -1.2655067  1.30444116  1.4142136 -1.12067661
#> [13,] -1.4591865  0.68112137  1.4142136 -1.69759214
#> [14,] -0.8366442  1.65874925  1.4142136 -0.85933881
#> [15,] -0.1034277 -0.40148668 -0.7071068  0.18108150
#> [16,] -0.1034277 -0.11935246 -0.7071068  0.09725617
#> [17,] -0.9196499 -0.51302811 -0.7071068 -1.17984744
#> [18,]  1.4736796 -1.95650552 -0.7071068 -0.16408163
#> [19,]  0.2424292 -1.30037942 -0.7071068 -0.55855379
#> [20,]  2.4559131 -1.89089291 -0.7071068  0.23039052
#> [21,] -0.6706329  0.73361146  1.4142136 -0.54869199
#> attr(,"center")
#>  [1]   0.4285714   3.0952381   6.3809524 245.7476190   3.5980952   3.7142857
#>  [7] 157.4761905  17.4819048   0.3333333   3.3363810
#> attr(,"scale")
#>  [1]   0.4948717   1.6876233   1.8120283 127.5760226   0.5511859   0.7648752
#>  [7]  72.2842399   1.5240973   0.4714045   1.0140133
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 35
#> 
#> $lambda.min
#> [1] 0.4605878
#> 
#> $null.dev
#> [1] 31.95333
#> 
#> $Bias
#> [1] 2.058516
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 11.90243