Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 48.631281 45.774102 40.895491 35.933708 31.892699 28.634033 26.038555
#>   [8] 24.003795 22.441711 21.276742 20.444113 19.888364 19.562079 19.424784
#>  [15] 19.441989 19.397632 18.813864 18.353391 18.276488 18.276488 18.276488
#>  [22] 18.297055 18.303874 18.269261 18.182076 18.030568 17.791838 16.436241
#>  [29] 13.299905 13.312678 13.358057 11.855990 10.783207 10.390354 10.220715
#>  [36] 10.094852 10.243456 10.302784 10.268881 10.208916 10.158751  8.280012
#>  [43]  8.181763  8.059433  7.784317  7.784390  7.840334  7.926270  7.926369
#>  [50]  7.835901 10.962433 11.116322 11.305306 11.115298 10.735249 11.157854
#>  [57] 10.924004 13.532389 13.533044 13.491196 13.699979 13.856715 13.939180
#>  [64] 13.705566 14.303529 14.303863 14.291284 14.278618 14.280609 14.158155
#>  [71] 13.985605 13.986658 13.810995 14.688945 14.639435 14.406935 14.407309
#>  [78] 14.668866 14.669277 14.843908 14.844363 14.899433 14.499544 14.499306
#>  [85] 14.499057 14.498800 14.498530 14.468719 14.468361 14.372616 14.372111
#>  [92] 14.402447 14.401897 14.401353 14.400816 14.400151 14.404131 14.403579
#>  [99] 14.410267 14.409852
#> 
#> $cvse
#>   [1] 12.363080 12.399855 11.572870 10.436716  9.411623  8.486043  7.653674
#>   [8]  6.912682  6.265029  5.715668  5.271306  4.938472  4.720866  4.616747
#>  [15]  4.617615  4.670562  4.582961  4.498434  4.482132  4.482132  4.482132
#>  [22]  4.478151  4.469753  4.436960  4.398557  4.341160  4.311893  4.378836
#>  [29]  3.845509  3.820601  3.794747  2.678172  2.516426  2.346051  2.345933
#>  [36]  2.320061  2.488721  2.498540  2.510222  2.508213  2.487786  2.051428
#>  [43]  2.067091  2.119852  2.128634  2.128641  2.089966  2.116652  2.116706
#>  [50]  2.127082  3.219901  3.344289  3.332491  3.087324  3.091504  3.066075
#>  [57]  3.044070  4.596906  4.597324  4.601576  4.610070  4.599468  4.605297
#>  [64]  4.627921  4.614612  4.614646  4.616407  4.555730  4.556649  4.543319
#>  [71]  4.475824  4.476524  4.481455  5.079813  5.080950  5.107877  5.108008
#>  [78]  5.081455  5.081541  5.151948  5.152108  5.152319  4.791441  4.791023
#>  [85]  4.790614  4.790214  4.789821  4.789924  4.789550  4.793018  4.792669
#>  [92]  4.791227  4.790894  4.790568  4.790249  4.789953  4.789449  4.789145
#>  [99]  4.790443  4.790153
#> 
#> $fold
#>  [1]  3  5  4  4  7  7  8  1 10 10  3  6  9  2  8  6  1  2  9  5  1
#> 
#> $lambda
#>   [1] 5.841484916 5.447788383 5.080625679 4.738208513 4.418869117 4.121052128
#>   [7] 3.843307007 3.584280978 3.342712436 3.117424806 2.907320808 2.711377116
#>  [13] 2.528639373 2.358217542 2.199281572 2.051057355 1.912822954 1.783905089
#>  [19] 1.663675856 1.551549672 1.446980417 1.349458780 1.258509774 1.173690428
#>  [25] 1.094587621 1.020816079 0.952016492 0.887853766 0.828015394 0.772209928
#>  [31] 0.720165563 0.671628815 0.626363280 0.584148492 0.544778838 0.508062568
#>  [37] 0.473820851 0.441886912 0.412105213 0.384330701 0.358428098 0.334271244
#>  [43] 0.311742481 0.290732081 0.271137712 0.252863937 0.235821754 0.219928157
#>  [49] 0.205105736 0.191282295 0.178390509 0.166367586 0.155154967 0.144698042
#>  [55] 0.134945878 0.125850977 0.117369043 0.109458762 0.102081608 0.095201649
#>  [61] 0.088785377 0.082801539 0.077220993 0.072016556 0.067162881 0.062636328
#>  [67] 0.058414849 0.054477884 0.050806257 0.047382085 0.044188691 0.041210521
#>  [73] 0.038433070 0.035842810 0.033427124 0.031174248 0.029073208 0.027113771
#>  [79] 0.025286394 0.023582175 0.021992816 0.020510574 0.019128230 0.017839051
#>  [85] 0.016636759 0.015515497 0.014469804 0.013494588 0.012585098 0.011736904
#>  [91] 0.010945876 0.010208161 0.009520165 0.008878538 0.008280154 0.007722099
#>  [97] 0.007201656 0.006716288 0.006263633 0.005841485
#> 
#> $fit
#> $beta
#>               5.8415     5.4478     5.0806     4.7382    4.4189   4.1211
#> (Intercept) 20.81429 22.7364847 24.5875740 26.3139061 27.923889 29.42536
#> am           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.00000
#> carb         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.00000
#> cyl          0.00000 -0.3105091 -0.6095312 -0.8884002 -1.148474 -1.39102
#> disp         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.00000
#> drat         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.00000
#> gear         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.00000
#> hp           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.00000
#> qsec         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.00000
#> vs           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.00000
#> wt           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.00000
#>               3.8433    3.5843    3.3427    3.1174    2.9073    2.7114
#> (Intercept) 30.82565 32.131553 33.349446 34.485257 35.544519 36.532389
#> am           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.61722 -1.828174 -2.024911 -2.208388 -2.379499 -2.539078
#> disp         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                2.5286    2.3582    2.1993    2.0511   1.9128   1.7839   1.6637
#> (Intercept) 37.453681 38.312880 39.114173 39.861461 40.40615 40.40615 40.40615
#> am           0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#> carb         0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#> cyl         -2.687902 -2.826696 -2.956136 -3.076851 -3.16484 -3.16484 -3.16484
#> disp         0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#> drat         0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#> gear         0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#> hp           0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#> vs           0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#> wt           0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#>               1.5515   1.4470   1.3495   1.2585   1.1737     1.0946    1.0208
#> (Intercept) 40.40615 40.40615 40.40615 40.40615 40.40615 40.0975399 39.359931
#> am           0.00000  0.00000  0.00000  0.00000  0.00000  0.2133458  0.723264
#> carb         0.00000  0.00000  0.00000  0.00000  0.00000  0.0000000  0.000000
#> cyl         -3.16484 -3.16484 -3.16484 -3.16484 -3.16484 -3.1297573 -3.045907
#> disp         0.00000  0.00000  0.00000  0.00000  0.00000  0.0000000  0.000000
#> drat         0.00000  0.00000  0.00000  0.00000  0.00000  0.0000000  0.000000
#> gear         0.00000  0.00000  0.00000  0.00000  0.00000  0.0000000  0.000000
#> hp           0.00000  0.00000  0.00000  0.00000  0.00000  0.0000000  0.000000
#> qsec         0.00000  0.00000  0.00000  0.00000  0.00000  0.0000000  0.000000
#> vs           0.00000  0.00000  0.00000  0.00000  0.00000  0.0000000  0.000000
#> wt           0.00000  0.00000  0.00000  0.00000  0.00000  0.0000000  0.000000
#>                0.9520    0.8879    0.8280      0.7722     0.7202    0.6716
#> (Intercept) 38.671332 22.726191 22.726190 22.71257849 22.1933951 21.709044
#> am           1.199301 10.678568 10.678571 10.67136824 10.3796808 10.107544
#> carb         0.000000 -2.390476 -2.390476 -2.38710054 -2.2529259 -2.127747
#> cyl         -2.967628  0.000000  0.000000  0.00000000  0.0000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.01978356  0.7547942  1.440471
#> wt           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.000000
#>                0.6264    0.5841    0.5448    0.5081    0.4738     0.4419
#> (Intercept) 21.258168 20.836853 20.443928 20.377527 20.377206 19.5261789
#> am           9.854312  9.617588  9.396813  9.359358  9.359061  9.1508188
#> carb        -2.011251 -1.902362 -1.800810 -1.783593 -1.783477 -1.7728263
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.2612554
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> vs           2.078871  2.675302  3.231545  3.325315  3.325661  3.2648088
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#>                0.4121    0.3843    0.3584    0.3343    0.3117    0.2907
#> (Intercept) 14.547689 13.403136 13.401870 13.401131 13.400703 13.400458
#> am           7.926953  7.645477  7.644735  7.644289  7.644022  7.643865
#> carb        -1.710055 -1.695804 -1.695640 -1.695536 -1.695472 -1.695433
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         1.789618  2.141189  2.141461  2.141613  2.141697  2.141743
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           2.911615  2.829945  2.830381  2.830659  2.830834  2.830943
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                0.2711    0.2529    0.2358    0.2199    0.2051    0.1913
#> (Intercept) 13.400319 13.400241 13.400198 13.400174 13.400161 10.152518
#> am           7.643773  7.643720  7.643689  7.643672  7.643662  6.118644
#> carb        -1.695409 -1.695394 -1.695386 -1.695380 -1.695377 -1.864038
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         2.141767  2.141780  2.141787  2.141790  2.141792  2.065005
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  1.291964
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           2.831010  2.831051  2.831075  2.831090  2.831099  2.559390
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                0.1784    0.1664    0.1552    0.1447    0.1349    0.1259
#> (Intercept) 10.151530 10.150778 10.150206 10.149771 10.149440 10.149188
#> am           6.117836  6.117219  6.116750  6.116392  6.116120  6.115913
#> carb        -1.864154 -1.864243 -1.864311 -1.864362 -1.864402 -1.864432
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         2.064839  2.064711  2.064613  2.064538  2.064481  2.064437
#> gear         1.292587  1.293063  1.293426  1.293703  1.293914  1.294075
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           2.559282  2.559199  2.559135  2.559087  2.559049  2.559021
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                  0.1174      0.1095      0.1021      0.0952      0.0888
#> (Intercept) -14.5672790 -14.5719000 -14.5761873 -14.5801654 -14.5838567
#> am            6.3201351   6.3202048   6.3202687   6.3203272   6.3203810
#> carb         -2.2894774  -2.2896048  -2.2897231  -2.2898329  -2.2899348
#> cyl           1.4147763   1.4150947   1.4153901   1.4156643   1.4159187
#> disp          0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> drat          3.3833038   3.3835003   3.3836828   3.3838523   3.3840097
#> gear          2.7666308   2.7669889   2.7673215   2.7676304   2.7679172
#> hp            0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> qsec          0.3624614   0.3625075   0.3625502   0.3625897   0.3626262
#> vs            3.7217415   3.7220767   3.7223880   3.7226773   3.7229459
#> wt            0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#>                  0.0828      0.0772      0.0720      0.0672      0.0626
#> (Intercept) -14.5872822 -14.5904612 -14.5934115 -14.5961498 -14.5986915
#> am            6.3204303   6.3204756   6.3205173   6.3205557   6.3205911
#> carb         -2.2900294  -2.2901172  -2.2901987  -2.2902744  -2.2903446
#> cyl           1.4161548   1.4163740   1.4165774   1.4167663   1.4169416
#> disp          0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> drat          3.3841560   3.3842918   3.3844180   3.3845351   3.3846440
#> gear          2.7681837   2.7684311   2.7686609   2.7688742   2.7690724
#> hp            0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> qsec          0.3626601   0.3626915   0.3627206   0.3627475   0.3627725
#> vs            3.7231955   3.7234273   3.7236425   3.7238425   3.7240282
#> wt            0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#>                  0.0584      0.0545     0.0508        0.0474        0.0442
#> (Intercept) -14.6010506 -14.6032404 -14.605273 -14.368473686 -14.372509638
#> am            6.3206237   6.3206537   6.320681   7.122738541   7.122515650
#> carb         -2.2904099  -2.2904704  -2.290527  -2.603059541  -2.603065790
#> cyl           1.4171043   1.4172553   1.417396   1.225058093   1.225392383
#> disp          0.0000000   0.0000000   0.000000  -0.004717412  -0.004715568
#> drat          3.3847451   3.3848389   3.384926   3.190764982   3.190993127
#> gear          2.7692565   2.7694274   2.769586   2.319237967   2.319711674
#> hp            0.0000000   0.0000000   0.000000   0.016063207   0.016058069
#> qsec          0.3627957   0.3628171   0.362837   0.526728582   0.526705184
#> vs            3.7242007   3.7243610   3.724510   2.728496541   2.729204611
#> wt            0.0000000   0.0000000   0.000000   0.000000000   0.000000000
#>                   0.0412        0.0384        0.0358        0.0334
#> (Intercept) -14.37628844 -14.379824279 -14.383130697 -14.386220608
#> am            7.12230566   7.122107941   7.121921870   7.121746858
#> carb         -2.60306994  -2.603072209  -2.603072798  -2.603071894
#> cyl           1.22570685   1.226002502   1.226280299   1.226541169
#> disp         -0.00471378  -0.004712051  -0.004710378  -0.004708764
#> drat          3.19120611   3.191404819   3.191590075   3.191762675
#> gear          2.32015900   2.320581177   2.320979381   2.321354757
#> hp            0.01605314   0.016048430   0.016043919   0.016039606
#> qsec          0.52668177   0.526658434   0.526635257   0.526612312
#> vs            2.72987963   2.730522686   2.731134865   2.731717256
#> wt            0.00000000   0.000000000   0.000000000   0.000000000
#>                    0.0312        0.0291        0.0271        0.0253
#> (Intercept) -14.389106324 -14.391799567 -14.394311496 -14.396652721
#> am            7.121582339   7.121427770   7.121282628   7.121146414
#> carb         -2.603069665  -2.603066270  -2.603061852  -2.603056542
#> cyl           1.226785999   1.227015641   1.227230909   1.227432582
#> disp         -0.004707207  -0.004705708  -0.004704265  -0.004702878
#> drat          3.191923370   3.192072875   3.192211867   3.192340990
#> gear          2.321708413   2.322041412   2.322354781   2.322649509
#> hp            0.016035486   0.016031554   0.016027802   0.016024227
#> qsec          0.526589661   0.526567359   0.526545454   0.526523987
#> vs            2.732270938   2.732796985   2.733296455   2.733770393
#> wt            0.000000000   0.000000000   0.000000000   0.000000000
#>                    0.0236        0.0220        0.0205       0.0191
#> (Intercept) -14.398833327 -14.400862891 -14.402750505 -14.40450479
#> am            7.121018652   7.120898883   7.120786673   7.12068160
#> carb         -2.603050460  -2.603043717  -2.603036412  -2.60302864
#> cyl           1.227621403   1.227798085   1.227963304   1.22811771
#> disp         -0.004701546  -0.004700268  -0.004699043  -0.00469787
#> drat          3.192460851   3.192572026   3.192675061   3.19277047
#> gear          2.322926544   2.323186798   2.323431146   2.32366043
#> hp            0.016020821   0.016017579   0.016014496   0.01601156
#> qsec          0.526502990   0.526482494   0.526462522   0.52644309
#> vs            2.734219825   2.734645758   2.735049176   2.73543104
#> wt            0.000000000   0.000000000   0.000000000   0.00000000
#>                    0.0178        0.0166        0.0155        0.0145
#> (Intercept) -14.406133916 -1.440816e+01 -13.842089355 -13.838561231
#> am            7.120583278  7.120192e+00   7.262251897   7.262751765
#> carb         -2.603020474 -2.602962e+00  -2.505676565  -2.505734571
#> cyl           1.228261914  1.228533e+00   1.118927872   1.118582097
#> disp         -0.004696748 -4.694545e-03  -0.001955759  -0.001957182
#> drat          3.192858745  3.193027e+00   3.022929847   3.022653275
#> gear          2.323875441  2.324260e+00   2.122079930   2.121599480
#> hp            0.016008780  1.600343e-02   0.016921692   0.016928379
#> qsec          0.526424219  5.263915e-01   0.631701731   0.631750442
#> vs            2.735792295  2.736443e+00   2.425532943   2.424782370
#> wt            0.000000000 -3.331824e-04  -0.441804756  -0.441780747
#>                    0.0135        0.0126        0.0117        0.0109
#> (Intercept) -13.835138533 -13.831819382 -13.828601904 -13.825484231
#> am            7.263234275   7.263699792   7.264148677   7.264581290
#> carb         -2.505792739  -2.505851023  -2.505909379  -2.505967763
#> cyl           1.118247625   1.117924230   1.117611692   1.117309788
#> disp         -0.001958626  -0.001960088  -0.001961567  -0.001963061
#> drat          3.022387645   3.022132702   3.021888190   3.021653861
#> gear          2.121135262   2.120686945   2.120254198   2.119836694
#> hp            0.016934880   0.016941198   0.016947335   0.016953296
#> qsec          0.631796142   0.631838920   0.631878863   0.631916059
#> vs            2.424056023   2.423353430   2.422674121   2.422017628
#> wt           -0.441750062  -0.441712990  -0.441669812  -0.441620802
#>                    0.0102        0.0095        0.0089        0.0083
#> (Intercept) -13.822464500 -13.819540855 -13.816711451 -13.813974450
#> am            7.264997986   7.265399118   7.265785034   7.266156079
#> carb         -2.506026135  -2.506084454  -2.506142683  -2.506200785
#> cyl           1.117018300   1.116737011   1.116465703   1.116204164
#> disp         -0.001964568  -0.001966087  -0.001967616  -0.001969155
#> drat          3.021429465   3.021214760   3.021009504   3.020813458
#> gear          2.119434109   2.119046120   2.118672409   2.118312658
#> hp            0.016959082   0.016964697   0.016970145   0.016975427
#> qsec          0.631950592   0.631982545   0.632011999   0.632039035
#> vs            2.421383487   2.420771236   2.420180419   2.419610581
#> wt           -0.441566229  -0.441506353  -0.441441427  -0.441371699
#>                   0.0077        0.0072        0.0067        0.0063
#> (Intercept) -13.81132803 -13.808770361 -13.806299653 -13.803914110
#> am            7.26651260   7.266854922   7.267183391   7.267498334
#> carb         -2.50625873  -2.506316473  -2.506373993  -2.506431257
#> cyl           1.11595218   1.115709543   1.115476043   1.115251474
#> disp         -0.00197070  -0.001972252  -0.001973809  -0.001975369
#> drat          3.02062639   3.020448063   3.020278254   3.020116736
#> gear          2.11796656   2.117633793   2.117314062   2.117007060
#> hp            0.01698055   0.016985507   0.016990312   0.016994964
#> qsec          0.63206373   0.632086162   0.632106406   0.632124536
#> vs            2.41906127   2.418532044   2.418022457   2.417532073
#> wt           -0.44129741  -0.441218790  -0.441136071  -0.441049472
#>                    0.0058
#> (Intercept) -13.801611954
#> am            7.267800076
#> carb         -2.506488234
#> cyl           1.115035633
#> disp         -0.001976931
#> drat          3.019963287
#> gear          2.116712488
#> hp            0.016999465
#> qsec          0.632140625
#> vs            2.417060459
#> wt           -0.440959208
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1   1  10  10  10  39   1   7  22  22  21  21  21   8
#>  [37]   1  47  77  13   1   1   1   1   1   1   1   1   1  32   1   1   1   1
#>  [55]   1   1 139   1   1   1   1   1   1   1   1   1   1   1   1 186   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   2 159   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.841484916 5.447788383 5.080625679 4.738208513 4.418869117 4.121052128
#>   [7] 3.843307007 3.584280978 3.342712436 3.117424806 2.907320808 2.711377116
#>  [13] 2.528639373 2.358217542 2.199281572 2.051057355 1.912822954 1.783905089
#>  [19] 1.663675856 1.551549672 1.446980417 1.349458780 1.258509774 1.173690428
#>  [25] 1.094587621 1.020816079 0.952016492 0.887853766 0.828015394 0.772209928
#>  [31] 0.720165563 0.671628815 0.626363280 0.584148492 0.544778838 0.508062568
#>  [37] 0.473820851 0.441886912 0.412105213 0.384330701 0.358428098 0.334271244
#>  [43] 0.311742481 0.290732081 0.271137712 0.252863937 0.235821754 0.219928157
#>  [49] 0.205105736 0.191282295 0.178390509 0.166367586 0.155154967 0.144698042
#>  [55] 0.134945878 0.125850977 0.117369043 0.109458762 0.102081608 0.095201649
#>  [61] 0.088785377 0.082801539 0.077220993 0.072016556 0.067162881 0.062636328
#>  [67] 0.058414849 0.054477884 0.050806257 0.047382085 0.044188691 0.041210521
#>  [73] 0.038433070 0.035842810 0.033427124 0.031174248 0.029073208 0.027113771
#>  [79] 0.025286394 0.023582175 0.021992816 0.020510574 0.019128230 0.017839051
#>  [85] 0.016636759 0.015515497 0.014469804 0.013494588 0.012585098 0.011736904
#>  [91] 0.010945876 0.010208161 0.009520165 0.008878538 0.008280154 0.007722099
#>  [97] 0.007201656 0.006716288 0.006263633 0.005841485
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 27
#> 
#> $loss
#>   [1] 925.20571 792.06122 676.82636 580.83838 501.30478 435.81601 382.29432
#>   [8] 338.94935 304.23953 276.83865 255.60680 239.56524 227.87450 219.81546
#>  [15] 214.77297 212.22155 211.67003 211.67003 211.67003 211.67003 211.67003
#>  [22] 211.67003 211.67003 211.67003 206.82517 196.43176 188.23863 102.68429
#>  [29] 102.68429 102.37292  92.12127  84.86933  80.12330  77.43698  76.45317
#>  [36]  76.43196  76.43196  73.81321  65.30947  65.00040  65.00039  65.00039
#>  [43]  65.00039  65.00039  65.00039  65.00039  65.00039  65.00039  65.00039
#>  [50]  60.75914  60.75913  60.75912  60.75912  60.75912  60.75912  60.75912
#>  [57]  53.58123  53.58122  53.58121  53.58120  53.58119  53.58119  53.58118
#>  [64]  53.58118  53.58117  53.58117  53.58117  53.58116  53.58116  52.04561
#>  [71]  52.04559  52.04558  52.04556  52.04555  52.04554  52.04553  52.04552
#>  [78]  52.04551  52.04550  52.04550  52.04549  52.04548  52.04548  52.04548
#>  [85]  52.04522  51.88540  51.88537  51.88535  51.88533  51.88532  51.88530
#>  [92]  51.88528  51.88527  51.88525  51.88524  51.88523  51.88522  51.88521
#>  [99]  51.88520  51.88519
#> 
#> $linear.predictors
#>      5.8415   5.4478   5.0806   4.7382   4.4189   4.1211   3.8433   3.5843
#> 1  20.81429 20.87343 20.93039 20.98350 21.03304 21.07924 21.12233 21.16251
#> 2  20.81429 20.25241 19.71132 19.20670 18.73609 18.29720 17.88789 17.50616
#> 3  20.81429 20.87343 20.93039 20.98350 21.03304 21.07924 21.12233 21.16251
#> 4  20.81429 21.49445 22.14945 22.76031 23.32999 23.86128 24.35677 24.81886
#> 5  20.81429 20.25241 19.71132 19.20670 18.73609 18.29720 17.88789 17.50616
#> 6  20.81429 20.25241 19.71132 19.20670 18.73609 18.29720 17.88789 17.50616
#> 7  20.81429 20.25241 19.71132 19.20670 18.73609 18.29720 17.88789 17.50616
#> 8  20.81429 20.25241 19.71132 19.20670 18.73609 18.29720 17.88789 17.50616
#> 9  20.81429 21.49445 22.14945 22.76031 23.32999 23.86128 24.35677 24.81886
#> 10 20.81429 21.49445 22.14945 22.76031 23.32999 23.86128 24.35677 24.81886
#> 11 20.81429 21.49445 22.14945 22.76031 23.32999 23.86128 24.35677 24.81886
#> 12 20.81429 21.49445 22.14945 22.76031 23.32999 23.86128 24.35677 24.81886
#> 13 20.81429 20.25241 19.71132 19.20670 18.73609 18.29720 17.88789 17.50616
#> 14 20.81429 20.25241 19.71132 19.20670 18.73609 18.29720 17.88789 17.50616
#> 15 20.81429 20.25241 19.71132 19.20670 18.73609 18.29720 17.88789 17.50616
#> 16 20.81429 20.25241 19.71132 19.20670 18.73609 18.29720 17.88789 17.50616
#> 17 20.81429 21.49445 22.14945 22.76031 23.32999 23.86128 24.35677 24.81886
#> 18 20.81429 21.49445 22.14945 22.76031 23.32999 23.86128 24.35677 24.81886
#> 19 20.81429 21.49445 22.14945 22.76031 23.32999 23.86128 24.35677 24.81886
#> 20 20.81429 20.87343 20.93039 20.98350 21.03304 21.07924 21.12233 21.16251
#> 21 20.81429 20.25241 19.71132 19.20670 18.73609 18.29720 17.88789 17.50616
#>      3.3427   3.1174   2.9073   2.7114   2.5286   2.3582   2.1993   2.0511
#> 1  21.19998 21.23493 21.26752 21.29792 21.32627 21.35270 21.37736 21.40035
#> 2  17.15016 16.81816 16.50853 16.21976 15.95046 15.69931 15.46509 15.24665
#> 3  21.19998 21.23493 21.26752 21.29792 21.32627 21.35270 21.37736 21.40035
#> 4  25.24980 25.65171 26.02652 26.37608 26.70207 27.00610 27.28963 27.55406
#> 5  17.15016 16.81816 16.50853 16.21976 15.95046 15.69931 15.46509 15.24665
#> 6  17.15016 16.81816 16.50853 16.21976 15.95046 15.69931 15.46509 15.24665
#> 7  17.15016 16.81816 16.50853 16.21976 15.95046 15.69931 15.46509 15.24665
#> 8  17.15016 16.81816 16.50853 16.21976 15.95046 15.69931 15.46509 15.24665
#> 9  25.24980 25.65171 26.02652 26.37608 26.70207 27.00610 27.28963 27.55406
#> 10 25.24980 25.65171 26.02652 26.37608 26.70207 27.00610 27.28963 27.55406
#> 11 25.24980 25.65171 26.02652 26.37608 26.70207 27.00610 27.28963 27.55406
#> 12 25.24980 25.65171 26.02652 26.37608 26.70207 27.00610 27.28963 27.55406
#> 13 17.15016 16.81816 16.50853 16.21976 15.95046 15.69931 15.46509 15.24665
#> 14 17.15016 16.81816 16.50853 16.21976 15.95046 15.69931 15.46509 15.24665
#> 15 17.15016 16.81816 16.50853 16.21976 15.95046 15.69931 15.46509 15.24665
#> 16 17.15016 16.81816 16.50853 16.21976 15.95046 15.69931 15.46509 15.24665
#> 17 25.24980 25.65171 26.02652 26.37608 26.70207 27.00610 27.28963 27.55406
#> 18 25.24980 25.65171 26.02652 26.37608 26.70207 27.00610 27.28963 27.55406
#> 19 25.24980 25.65171 26.02652 26.37608 26.70207 27.00610 27.28963 27.55406
#> 20 21.19998 21.23493 21.26752 21.29792 21.32627 21.35270 21.37736 21.40035
#> 21 17.15016 16.81816 16.50853 16.21976 15.95046 15.69931 15.46509 15.24665
#>      1.9128   1.7839   1.6637   1.5515   1.4470   1.3495   1.2585   1.1737
#> 1  21.41711 21.41711 21.41711 21.41711 21.41711 21.41711 21.41711 21.41711
#> 2  15.08743 15.08743 15.08743 15.08743 15.08743 15.08743 15.08743 15.08743
#> 3  21.41711 21.41711 21.41711 21.41711 21.41711 21.41711 21.41711 21.41711
#> 4  27.74679 27.74679 27.74679 27.74679 27.74679 27.74679 27.74679 27.74679
#> 5  15.08743 15.08743 15.08743 15.08743 15.08743 15.08743 15.08743 15.08743
#> 6  15.08743 15.08743 15.08743 15.08743 15.08743 15.08743 15.08743 15.08743
#> 7  15.08743 15.08743 15.08743 15.08743 15.08743 15.08743 15.08743 15.08743
#> 8  15.08743 15.08743 15.08743 15.08743 15.08743 15.08743 15.08743 15.08743
#> 9  27.74679 27.74679 27.74679 27.74679 27.74679 27.74679 27.74679 27.74679
#> 10 27.74679 27.74679 27.74679 27.74679 27.74679 27.74679 27.74679 27.74679
#> 11 27.74679 27.74679 27.74679 27.74679 27.74679 27.74679 27.74679 27.74679
#> 12 27.74679 27.74679 27.74679 27.74679 27.74679 27.74679 27.74679 27.74679
#> 13 15.08743 15.08743 15.08743 15.08743 15.08743 15.08743 15.08743 15.08743
#> 14 15.08743 15.08743 15.08743 15.08743 15.08743 15.08743 15.08743 15.08743
#> 15 15.08743 15.08743 15.08743 15.08743 15.08743 15.08743 15.08743 15.08743
#> 16 15.08743 15.08743 15.08743 15.08743 15.08743 15.08743 15.08743 15.08743
#> 17 27.74679 27.74679 27.74679 27.74679 27.74679 27.74679 27.74679 27.74679
#> 18 27.74679 27.74679 27.74679 27.74679 27.74679 27.74679 27.74679 27.74679
#> 19 27.74679 27.74679 27.74679 27.74679 27.74679 27.74679 27.74679 27.74679
#> 20 21.41711 21.41711 21.41711 21.41711 21.41711 21.41711 21.41711 21.41711
#> 21 15.08743 15.08743 15.08743 15.08743 15.08743 15.08743 15.08743 15.08743
#>      1.0946   1.0208   0.9520   0.8879   0.8280   0.7722   0.7202   0.6716
#> 1  21.53234 21.80775 22.06486 23.84286 23.84286 23.83554 23.56137 23.30560
#> 2  15.05948 14.99267 14.93031 17.94524 17.94524 17.93838 17.68754 17.45355
#> 3  21.31900 21.08449 20.86556 20.33572 20.33571 20.34526 20.69526 21.02177
#> 4  27.57851 27.17630 26.80082 17.94524 17.94524 17.95816 18.44234 18.89402
#> 5  15.05948 14.99267 14.93031 15.55476 15.55476 15.55128 15.43462 15.32580
#> 6  15.05948 14.99267 14.93031 15.55476 15.55476 15.55128 15.43462 15.32580
#> 7  15.05948 14.99267 14.93031 13.16429 13.16429 13.16418 13.18169 13.19806
#> 8  15.05948 14.99267 14.93031 13.16429 13.16429 13.16418 13.18169 13.19806
#> 9  27.79186 27.89957 28.00012 31.01428 31.01429 31.01663 31.07494 31.12931
#> 10 27.79186 27.89957 28.00012 28.62381 28.62381 28.62953 28.82202 29.00157
#> 11 27.79186 27.89957 28.00012 31.01428 31.01429 31.01663 31.07494 31.12931
#> 12 27.57851 27.17630 26.80082 20.33572 20.33571 20.34526 20.69526 21.02177
#> 13 15.05948 14.99267 14.93031 17.94524 17.94524 17.93838 17.68754 17.45355
#> 14 15.05948 14.99267 14.93031 17.94524 17.94524 17.93838 17.68754 17.45355
#> 15 15.05948 14.99267 14.93031 13.16429 13.16429 13.16418 13.18169 13.19806
#> 16 15.05948 14.99267 14.93031 17.94524 17.94524 17.93838 17.68754 17.45355
#> 17 27.79186 27.89957 28.00012 31.01428 31.01429 31.01663 31.07494 31.12931
#> 18 27.79186 27.89957 28.00012 28.62381 28.62381 28.60975 28.06722 27.56109
#> 19 27.79186 27.89957 28.00012 28.62381 28.62381 28.62953 28.82202 29.00157
#> 20 21.53234 21.80775 22.06486 19.06190 19.06190 19.06134 19.05552 19.05010
#> 21 15.27283 15.71594 16.12961 14.28095 14.28095 14.28714 14.54967 14.79461
#>      0.6264   0.5841   0.5448   0.5081   0.4738   0.4419   0.4121   0.3843
#> 1  23.06748 22.84499 22.63750 22.60251 22.60236 22.60459 22.61394 22.61603
#> 2  17.23567 17.03213 16.84231 16.81034 16.81025 16.80348 16.76488 16.75627
#> 3  21.32579 21.60979 21.87466 21.91925 21.91939 21.73923 20.68860 20.44696
#> 4  19.31454 19.70743 20.07385 20.13566 20.13591 20.26946 21.05450 21.23493
#> 5  15.22442 15.12977 15.04150 15.02675 15.02678 15.00975 14.91165 14.88917
#> 6  15.22442 15.12977 15.04150 15.02675 15.02678 15.00975 14.91165 14.88917
#> 7  13.21316 13.22741 13.24069 13.24315 13.24330 13.20035 12.95105 12.89360
#> 8  13.21316 13.22741 13.24069 13.24315 13.24330 13.27873 13.48794 13.53596
#> 9  31.18010 31.22738 31.27148 31.27861 31.27845 31.23490 30.97785 30.91881
#> 10 29.16885 29.32502 29.47067 29.49501 29.49498 29.68414 30.78897 31.04301
#> 11 31.18010 31.22738 31.27148 31.27861 31.27845 31.27148 31.22839 31.21857
#> 12 21.32579 21.60979 21.87466 21.91925 21.91939 21.98481 22.37084 22.45968
#> 13 17.23567 17.03213 16.84231 16.81034 16.81025 16.70159 16.06693 15.92121
#> 14 17.23567 17.03213 16.84231 16.81034 16.81025 16.80348 16.76488 16.75627
#> 15 13.21316 13.22741 13.24069 13.24315 13.24330 13.40936 14.38275 14.60655
#> 16 17.23567 17.03213 16.84231 16.81034 16.81025 16.78519 16.63960 16.60639
#> 17 31.18010 31.22738 31.27148 31.27861 31.27845 31.23490 30.97785 30.91881
#> 18 27.08998 26.64972 26.23912 26.16970 26.16931 26.28871 26.98254 27.14247
#> 19 29.16885 29.32502 29.47067 29.49501 29.49498 29.38109 28.71301 28.55923
#> 20 19.04498 19.04027 19.03588 19.03533 19.03541 18.98578 18.69273 18.62489
#> 21 15.02247 15.23555 15.43426 15.46814 15.46845 15.41923 15.12945 15.06199
#>      0.3584   0.3343   0.3117   0.2907   0.2711   0.2529   0.2358   0.2199
#> 1  22.61574 22.61556 22.61546 22.61539 22.61535 22.61533 22.61531 22.61530
#> 2  16.75619 16.75614 16.75610 16.75608 16.75607 16.75606 16.75605 16.75605
#> 3  20.44704 20.44710 20.44715 20.44718 20.44720 20.44721 20.44722 20.44722
#> 4  21.23550 21.23584 21.23604 21.23617 21.23624 21.23628 21.23631 21.23632
#> 5  14.88924 14.88927 14.88930 14.88931 14.88932 14.88932 14.88933 14.88933
#> 6  14.88924 14.88927 14.88930 14.88931 14.88932 14.88932 14.88933 14.88933
#> 7  12.89379 12.89391 12.89399 12.89403 12.89406 12.89408 12.89409 12.89410
#> 8  13.53623 13.53639 13.53650 13.53656 13.53659 13.53661 13.53663 13.53663
#> 9  30.91851 30.91832 30.91821 30.91814 30.91810 30.91808 30.91807 30.91806
#> 10 31.04311 31.04316 31.04318 31.04319 31.04320 31.04320 31.04320 31.04320
#> 11 31.21831 31.21815 31.21805 31.21799 31.21795 31.21793 31.21792 31.21791
#> 12 22.46002 22.46022 22.46034 22.46042 22.46046 22.46048 22.46050 22.46051
#> 13 15.92102 15.92091 15.92084 15.92080 15.92078 15.92077 15.92076 15.92075
#> 14 16.75619 16.75614 16.75610 16.75608 16.75607 16.75606 16.75605 16.75605
#> 15 14.60696 14.60720 14.60734 14.60743 14.60748 14.60750 14.60752 14.60753
#> 16 16.60629 16.60623 16.60619 16.60616 16.60614 16.60614 16.60613 16.60613
#> 17 30.91851 30.91832 30.91821 30.91814 30.91810 30.91808 30.91807 30.91806
#> 18 27.14200 27.14169 27.14150 27.14138 27.14130 27.14126 27.14123 27.14121
#> 19 28.55901 28.55889 28.55881 28.55877 28.55875 28.55873 28.55873 28.55872
#> 20 18.62485 18.62484 18.62484 18.62484 18.62484 18.62484 18.62484 18.62484
#> 21 15.06226 15.06244 15.06256 15.06263 15.06268 15.06271 15.06273 15.06274
#>      0.2051   0.1913   0.1784   0.1664   0.1552   0.1447   0.1349   0.1259
#> 1  22.61530 22.03638 22.03597 22.03565 22.03541 22.03522 22.03508 22.03498
#> 2  16.75605 16.80510 16.80522 16.80532 16.80539 16.80545 16.80549 16.80553
#> 3  20.44723 20.42317 20.42337 20.42353 20.42364 20.42373 20.42380 20.42385
#> 4  21.23633 22.24651 22.24702 22.24741 22.24771 22.24793 22.24811 22.24824
#> 5  14.88933 14.77586 14.77588 14.77590 14.77591 14.77593 14.77593 14.77594
#> 6  14.88933 14.77586 14.77588 14.77590 14.77591 14.77593 14.77593 14.77594
#> 7  12.89410 12.62272 12.62265 12.62260 12.62256 12.62253 12.62250 12.62249
#> 8  13.53664 13.24222 13.24210 13.24201 13.24194 13.24189 13.24185 13.24182
#> 9  30.91805 30.55959 30.55938 30.55923 30.55911 30.55902 30.55895 30.55889
#> 10 31.04320 30.45081 30.45034 30.44999 30.44972 30.44951 30.44935 30.44923
#> 11 31.21790 30.84869 30.84846 30.84829 30.84815 30.84805 30.84797 30.84791
#> 12 22.46051 22.36428 22.36432 22.36435 22.36438 22.36440 22.36441 22.36442
#> 13 15.92075 15.99975 15.99994 16.00008 16.00020 16.00028 16.00035 16.00040
#> 14 16.75605 16.80510 16.80522 16.80532 16.80539 16.80545 16.80549 16.80553
#> 15 14.60753 14.27472 14.27452 14.27437 14.27425 14.27416 14.27409 14.27404
#> 16 16.60612 16.66055 16.66069 16.66079 16.66087 16.66093 16.66098 16.66102
#> 17 30.91805 30.55959 30.55938 30.55923 30.55911 30.55902 30.55895 30.55889
#> 18 27.14120 28.15088 28.15123 28.15150 28.15170 28.15186 28.15198 28.15207
#> 19 28.55872 29.34736 29.34772 29.34799 29.34819 29.34835 29.34847 29.34856
#> 20 18.62484 19.02207 19.02209 19.02211 19.02212 19.02213 19.02214 19.02215
#> 21 15.06275 15.12879 15.12860 15.12845 15.12833 15.12824 15.12818 15.12813
#>      0.1174   0.1095   0.1021   0.0952   0.0888   0.0828   0.0772   0.0720
#> 1  21.31113 21.31093 21.31075 21.31059 21.31043 21.31029 21.31015 21.31003
#> 2  17.29837 17.29852 17.29866 17.29878 17.29890 17.29902 17.29912 17.29921
#> 3  20.32042 20.32047 20.32051 20.32055 20.32059 20.32062 20.32065 20.32068
#> 4  22.86405 22.86404 22.86404 22.86403 22.86402 22.86402 22.86401 22.86401
#> 5  14.87596 14.87599 14.87601 14.87603 14.87605 14.87606 14.87608 14.87610
#> 6  15.09344 15.09349 15.09354 15.09358 15.09362 15.09366 15.09370 15.09373
#> 7  12.32305 12.32295 12.32285 12.32276 12.32267 12.32260 12.32252 12.32246
#> 8  13.13506 13.13499 13.13492 13.13486 13.13481 13.13475 13.13470 13.13466
#> 9  30.77175 30.77181 30.77187 30.77192 30.77197 30.77202 30.77206 30.77210
#> 10 31.01374 31.01380 31.01385 31.01390 31.01395 31.01399 31.01403 31.01407
#> 11 31.40127 31.40138 31.40148 31.40158 31.40166 31.40174 31.40182 31.40189
#> 12 20.59506 20.59464 20.59426 20.59390 20.59357 20.59326 20.59297 20.59271
#> 13 15.92451 15.92458 15.92464 15.92469 15.92475 15.92480 15.92484 15.92488
#> 14 17.39986 17.40002 17.40017 17.40031 17.40044 17.40056 17.40067 17.40078
#> 15 14.09817 14.09810 14.09804 14.09798 14.09793 14.09788 14.09784 14.09780
#> 16 17.07241 17.07255 17.07267 17.07279 17.07290 17.07300 17.07310 17.07319
#> 17 30.56515 30.56518 30.56522 30.56525 30.56527 30.56530 30.56532 30.56535
#> 18 27.70730 27.70720 27.70711 27.70702 27.70694 27.70686 27.70679 27.70673
#> 19 29.26855 29.26867 29.26877 29.26887 29.26896 29.26905 29.26913 29.26920
#> 20 18.20351 18.20333 18.20315 18.20299 18.20284 18.20270 18.20257 18.20244
#> 21 15.85723 15.85737 15.85749 15.85761 15.85772 15.85782 15.85791 15.85800
#>      0.0672   0.0626   0.0584   0.0545   0.0508   0.0474   0.0442   0.0412
#> 1  21.30991 21.30981 21.30971 21.30961 21.30953 21.09543 21.09528 21.09514
#> 2  17.29930 17.29939 17.29946 17.29953 17.29960 17.31221 17.31234 17.31247
#> 3  20.32071 20.32074 20.32076 20.32078 20.32080 20.14721 20.14733 20.14745
#> 4  22.86400 22.86400 22.86399 22.86399 22.86399 22.76276 22.76279 22.76280
#> 5  14.87611 14.87612 14.87613 14.87615 14.87616 15.13157 15.13148 15.13140
#> 6  15.09376 15.09379 15.09381 15.09384 15.09386 15.44760 15.44751 15.44741
#> 7  12.32239 12.32234 12.32228 12.32223 12.32219 11.86333 11.86343 11.86353
#> 8  13.13462 13.13458 13.13454 13.13451 13.13447 13.07813 13.07812 13.07812
#> 9  30.77213 30.77217 30.77220 30.77223 30.77225 31.01952 31.01948 31.01944
#> 10 31.01410 31.01414 31.01417 31.01419 31.01422 30.41749 30.41772 30.41794
#> 11 31.40195 31.40201 31.40206 31.40212 31.40216 31.71251 31.71248 31.71245
#> 12 20.59246 20.59224 20.59202 20.59183 20.59165 20.95215 20.95167 20.95122
#> 13 15.92492 15.92496 15.92499 15.92502 15.92505 15.78535 15.78545 15.78554
#> 14 17.40087 17.40096 17.40104 17.40112 17.40119 17.32229 17.32244 17.32258
#> 15 14.09776 14.09773 14.09769 14.09766 14.09764 14.28030 14.28021 14.28013
#> 16 17.07327 17.07334 17.07341 17.07348 17.07354 16.91596 16.91615 16.91633
#> 17 30.56537 30.56539 30.56540 30.56542 30.56544 30.71787 30.71784 30.71782
#> 18 27.70667 27.70661 27.70656 27.70651 27.70647 27.87027 27.87008 27.86990
#> 19 29.26927 29.26933 29.26939 29.26944 29.26949 29.07048 29.07068 29.07087
#> 20 18.20233 18.20223 18.20213 18.20204 18.20196 17.92434 17.92425 17.92417
#> 21 15.85808 15.85816 15.85822 15.85829 15.85835 16.27322 16.27326 16.27329
#>      0.0384   0.0358   0.0334   0.0312   0.0291   0.0271   0.0253   0.0236
#> 1  21.09501 21.09489 21.09477 21.09466 21.09456 21.09447 21.09438 21.09430
#> 2  17.31258 17.31269 17.31280 17.31290 17.31299 17.31308 17.31316 17.31324
#> 3  20.14757 20.14768 20.14779 20.14789 20.14799 20.14809 20.14818 20.14826
#> 4  22.76282 22.76283 22.76284 22.76284 22.76285 22.76285 22.76285 22.76284
#> 5  15.13133 15.13125 15.13117 15.13110 15.13103 15.13096 15.13090 15.13083
#> 6  15.44732 15.44723 15.44714 15.44706 15.44697 15.44689 15.44681 15.44674
#> 7  11.86363 11.86372 11.86382 11.86392 11.86401 11.86411 11.86420 11.86429
#> 8  13.07811 13.07812 13.07812 13.07812 13.07813 13.07814 13.07815 13.07815
#> 9  31.01940 31.01936 31.01932 31.01929 31.01925 31.01922 31.01919 31.01916
#> 10 30.41815 30.41835 30.41854 30.41872 30.41889 30.41905 30.41920 30.41934
#> 11 31.71242 31.71239 31.71236 31.71233 31.71230 31.71227 31.71224 31.71221
#> 12 20.95079 20.95039 20.95001 20.94965 20.94931 20.94898 20.94868 20.94840
#> 13 15.78563 15.78572 15.78580 15.78587 15.78595 15.78602 15.78608 15.78615
#> 14 17.32271 17.32284 17.32295 17.32306 17.32316 17.32325 17.32334 17.32342
#> 15 14.28004 14.27997 14.27989 14.27982 14.27975 14.27969 14.27963 14.27957
#> 16 16.91650 16.91667 16.91682 16.91697 16.91711 16.91725 16.91738 16.91750
#> 17 30.71779 30.71776 30.71774 30.71772 30.71770 30.71768 30.71766 30.71765
#> 18 27.86972 27.86955 27.86939 27.86924 27.86910 27.86896 27.86883 27.86870
#> 19 29.07106 29.07123 29.07140 29.07156 29.07172 29.07186 29.07200 29.07214
#> 20 17.92409 17.92402 17.92395 17.92388 17.92382 17.92377 17.92372 17.92367
#> 21 16.27331 16.27334 16.27336 16.27338 16.27340 16.27342 16.27344 16.27345
#>      0.0220   0.0205   0.0191   0.0178   0.0166   0.0155   0.0145   0.0135
#> 1  21.09422 21.09415 21.09408 21.09402 21.09401 21.17752 21.17761 21.17770
#> 2  17.31331 17.31338 17.31344 17.31350 17.31360 17.47543 17.47533 17.47523
#> 3  20.14835 20.14843 20.14850 20.14858 20.14862 20.08196 20.08185 20.08174
#> 4  22.76284 22.76283 22.76283 22.76282 22.76285 22.79248 22.79238 22.79228
#> 5  15.13077 15.13071 15.13066 15.13060 15.13031 14.93891 14.93896 14.93902
#> 6  15.44666 15.44659 15.44652 15.44646 15.44624 15.44605 15.44613 15.44620
#> 7  11.86438 11.86446 11.86454 11.86462 11.86425 11.89440 11.89438 11.89436
#> 8  13.07817 13.07818 13.07819 13.07820 13.07766 12.89118 12.89127 12.89135
#> 9  31.01913 31.01911 31.01908 31.01906 31.01911 30.92778 30.92790 30.92801
#> 10 30.41947 30.41960 30.41972 30.41983 30.42038 30.41890 30.41857 30.41825
#> 11 31.71219 31.71216 31.71213 31.71211 31.71228 31.78182 31.78192 31.78201
#> 12 20.94813 20.94787 20.94763 20.94741 20.94722 21.06238 21.06276 21.06313
#> 13 15.78620 15.78626 15.78631 15.78636 15.78642 15.82548 15.82538 15.82528
#> 14 17.32350 17.32357 17.32364 17.32370 17.32380 17.34099 17.34082 17.34066
#> 15 14.27951 14.27946 14.27940 14.27936 14.27919 14.22769 14.22773 14.22777
#> 16 16.91761 16.91772 16.91782 16.91792 16.91796 17.02561 17.02549 17.02537
#> 17 30.71763 30.71761 30.71760 30.71759 30.71775 30.68420 30.68428 30.68436
#> 18 27.86858 27.86847 27.86836 27.86826 27.86823 27.79505 27.79526 27.79546
#> 19 29.07226 29.07238 29.07250 29.07260 29.07314 29.05037 29.05018 29.05000
#> 20 17.92362 17.92358 17.92354 17.92351 17.92355 17.89836 17.89835 17.89834
#> 21 16.27346 16.27347 16.27348 16.27349 16.27343 16.36343 16.36346 16.36348
#>      0.0126   0.0117   0.0109   0.0102   0.0095   0.0089   0.0083   0.0077
#> 1  21.17779 21.17787 21.17795 21.17802 21.17809 21.17816 21.17823 21.17829
#> 2  17.47513 17.47504 17.47494 17.47484 17.47474 17.47465 17.47455 17.47445
#> 3  20.08164 20.08154 20.08144 20.08135 20.08126 20.08117 20.08109 20.08101
#> 4  22.79219 22.79209 22.79200 22.79192 22.79183 22.79175 22.79167 22.79160
#> 5  14.93908 14.93914 14.93920 14.93927 14.93933 14.93940 14.93947 14.93954
#> 6  15.44628 15.44635 15.44642 15.44649 15.44656 15.44663 15.44669 15.44676
#> 7  11.89433 11.89431 11.89428 11.89425 11.89422 11.89419 11.89416 11.89413
#> 8  12.89143 12.89152 12.89160 12.89168 12.89177 12.89185 12.89193 12.89201
#> 9  30.92812 30.92824 30.92835 30.92845 30.92856 30.92866 30.92877 30.92887
#> 10 30.41794 30.41765 30.41735 30.41707 30.41680 30.41653 30.41627 30.41602
#> 11 31.78209 31.78218 31.78226 31.78234 31.78241 31.78248 31.78255 31.78261
#> 12 21.06350 21.06385 21.06419 21.06451 21.06483 21.06514 21.06544 21.06573
#> 13 15.82518 15.82509 15.82499 15.82490 15.82481 15.82472 15.82463 15.82455
#> 14 17.34049 17.34034 17.34018 17.34003 17.33989 17.33975 17.33961 17.33948
#> 15 14.22781 14.22785 14.22789 14.22793 14.22797 14.22802 14.22806 14.22810
#> 16 17.02524 17.02512 17.02500 17.02488 17.02476 17.02465 17.02453 17.02441
#> 17 30.68444 30.68452 30.68459 30.68467 30.68474 30.68481 30.68488 30.68494
#> 18 27.79565 27.79585 27.79603 27.79622 27.79640 27.79658 27.79675 27.79693
#> 19 29.04982 29.04965 29.04948 29.04932 29.04916 29.04901 29.04886 29.04871
#> 20 17.89833 17.89832 17.89831 17.89830 17.89830 17.89829 17.89829 17.89828
#> 21 16.36350 16.36352 16.36353 16.36355 16.36356 16.36357 16.36358 16.36358
#>      0.0072   0.0067   0.0063   0.0058
#> 1  21.17835 21.17841 21.17846 21.17851
#> 2  17.47435 17.47426 17.47416 17.47406
#> 3  20.08093 20.08085 20.08078 20.08071
#> 4  22.79152 22.79145 22.79138 22.79132
#> 5  14.93962 14.93969 14.93977 14.93984
#> 6  15.44682 15.44688 15.44694 15.44700
#> 7  11.89409 11.89406 11.89402 11.89398
#> 8  12.89209 12.89217 12.89225 12.89233
#> 9  30.92897 30.92906 30.92916 30.92925
#> 10 30.41578 30.41555 30.41532 30.41510
#> 11 31.78267 31.78273 31.78279 31.78284
#> 12 21.06601 21.06628 21.06654 21.06679
#> 13 15.82446 15.82438 15.82430 15.82422
#> 14 17.33935 17.33923 17.33911 17.33899
#> 15 14.22814 14.22818 14.22823 14.22827
#> 16 17.02430 17.02419 17.02408 17.02396
#> 17 30.68501 30.68507 30.68513 30.68519
#> 18 27.79709 27.79726 27.79742 27.79758
#> 19 29.04857 29.04844 29.04831 29.04818
#> 20 17.89828 17.89828 17.89828 17.89828
#> 21 16.36359 16.36359 16.36359 16.36359
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 18.7 18.1 22.8 16.4 15.2 10.4 14.7 32.4 30.4 33.9 21.5 15.5 15.2 13.3
#> [16] 19.2 27.3 26.0 30.4 19.7 15.0
#> 
#> $X
#>               am       carb        cyl        disp        drat       gear
#>  [1,]  1.1547005  0.7348469 -0.1034175 -0.53972437  0.56827105  0.4291975
#>  [2,] -0.8660254 -0.4082483  0.9824666  1.02642345 -0.75182763 -0.8583951
#>  [3,] -0.8660254 -0.9797959 -0.1034175 -0.03072633 -1.43827895 -0.8583951
#>  [4,] -0.8660254 -0.4082483 -1.1893017 -0.69007456  0.60347369  0.4291975
#>  [5,] -0.8660254  0.1632993  0.9824666  0.36707522 -0.89263816 -0.8583951
#>  [6,] -0.8660254  0.1632993  0.9824666  0.36707522 -0.89263816 -0.8583951
#>  [7,] -0.8660254  0.7348469  0.9824666  1.90346624 -1.13905658 -0.8583951
#>  [8,] -0.8660254  0.7348469  0.9824666  1.65288259 -0.61101711 -0.8583951
#>  [9,]  1.1547005 -0.9797959 -1.1893017 -1.17636346  0.88509474  0.4291975
#> [10,]  1.1547005 -0.4082483 -1.1893017 -1.19985568  2.38120658  0.4291975
#> [11,]  1.1547005 -0.9797959 -1.1893017 -1.23587708  1.13151316  0.4291975
#> [12,] -0.8660254 -0.9797959 -1.1893017 -0.85217086  0.21624474 -0.8583951
#> [13,] -0.8660254 -0.4082483  0.9824666  0.69753241 -1.43827895 -0.8583951
#> [14,] -0.8660254 -0.4082483  0.9824666  0.58790206 -0.75182763 -0.8583951
#> [15,] -0.8660254  0.7348469  0.9824666  0.94811606  0.26904868 -0.8583951
#> [16,] -0.8660254 -0.4082483  0.9824666  1.33965302 -0.87503684 -0.8583951
#> [17,]  1.1547005 -0.9797959 -1.1893017 -1.17401424  0.88509474  0.4291975
#> [18,]  1.1547005 -0.4082483 -1.1893017 -0.85060472  1.50114079  1.7167902
#> [19,]  1.1547005 -0.4082483 -1.1893017 -1.04793934  0.33945395  1.7167902
#> [20,]  1.1547005  1.8779421 -0.1034175 -0.65718546  0.07543421  1.7167902
#> [21,]  1.1547005  3.0210373  0.9824666  0.56440985 -0.06537632  1.7167902
#>                hp        qsec         vs          wt
#>  [1,] -0.51443524 -0.73547822 -0.7844645 -0.48980046
#>  [2,]  0.42183689 -0.43588612 -0.7844645  0.29275325
#>  [3,] -0.58645617  1.27606872  1.2747549  0.31183993
#>  [4,] -0.73049804  2.70983090  1.2747549  0.01599645
#>  [5,]  0.49385783 -0.23259148 -0.7844645  0.89398354
#>  [6,]  0.49385783  0.08840005 -0.7844645  0.61722674
#>  [7,]  0.85396249  0.07770033 -0.7844645  2.02009741
#>  [8,]  1.21406716 -0.22189177 -0.7844645  2.11075912
#>  [9,] -1.14821945  0.87482931  1.2747549 -0.89062065
#> [10,] -1.34987806  0.36659271  1.2747549 -1.44890591
#> [11,] -1.16262364  1.10487324  1.2747549 -1.23895248
#> [12,] -0.70168966  1.16372168  1.2747549 -0.63772219
#> [13,]  0.06173223 -0.51613400 -0.7844645  0.36909995
#> [14,]  0.06173223 -0.28609007 -0.7844645  0.28798158
#> [15,]  1.43012996 -1.29721340 -0.7844645  0.67448677
#> [16,]  0.42183689 -0.41983654 -0.7844645  0.67925844
#> [17,] -1.14821945  0.56988735  1.2747549 -1.14351910
#> [18,] -0.78811478 -0.60708161 -0.7844645 -0.94788068
#> [19,] -0.47122268 -0.50008443  1.2747549 -1.54624796
#> [20,]  0.42183689 -1.24906467 -0.7844645 -0.34665039
#> [21,]  2.72650676 -1.73055197 -0.7844645  0.41681664
#> attr(,"center")
#>  [1]   0.4285714   2.7142857   6.1904762 228.9238095   3.5771429   3.6666667
#>  [7] 145.7142857  17.8347619   0.3809524   3.1332381
#> attr(,"scale")
#>  [1]   0.4948717   1.7496355   1.8418171 127.7018660   0.5681393   0.7766432
#>  [7]  69.4242601   1.8692082   0.4856209   1.0478514
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 45
#> 
#> $lambda.min
#> [1] 0.2711377
#> 
#> $null.dev
#> [1] 44.05741
#> 
#> $Bias
#> [1] 3.847103
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 18.90471