Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 28.809436 27.064524 24.781208 22.200785 19.998414 18.185171 16.703479
#>   [8] 15.503726 14.543193 13.785133 13.197958 12.754548 12.431642 12.209317
#>  [15] 12.070526 11.961966 11.640257 11.399141 11.294475 11.385938 11.482649
#>  [22] 11.578885 11.615998 11.617885 11.628388 11.592598 11.505261 11.107503
#>  [29] 10.243745  9.979367 10.988858 16.279680 17.213614 17.483635 17.266716
#>  [36] 17.166973 16.830353 16.811152 16.760094 16.472874 16.460260 16.183742
#>  [43] 15.992390 15.307956 15.273634 15.273496 15.306304 15.372017 16.872091
#>  [50] 19.936761 19.933147 20.449652 20.449917 21.067591 21.619988 21.620834
#>  [57] 22.696818 22.587566 22.588412 22.589233 22.590031 23.499601 24.191494
#>  [64] 24.319823 24.320926 24.374091 24.336850 24.375887 24.379299 24.388765
#>  [71] 24.381395 24.382055 24.382690 24.383304 24.383900 24.384479 24.385044
#>  [78] 24.385596 24.660385 24.659755 24.659145 24.658555 24.691568 24.678659
#>  [85] 24.685717 24.685339 24.684973 24.684619 24.684277 24.683946 24.678003
#>  [92] 24.650132 24.655067 24.654781 24.654483 24.648650 24.648416 24.648370
#>  [99] 24.654615 24.654317
#> 
#> $cvse
#>   [1] 10.633215 10.596197 10.424638  9.632959  8.858879  8.162793  7.537952
#>   [8]  6.978770  6.480592  6.039487  5.652045  5.315199  5.026050  4.781719
#>  [15]  4.579228  4.402345  4.214766  4.056905  3.938294  3.930130  3.916759
#>  [22]  3.902953  3.823593  3.741014  3.660653  3.545982  3.440245  3.359350
#>  [29]  3.255311  3.161071  3.280094  7.244743  8.216059  8.668290  8.641298
#>  [36]  8.631118  8.644486  8.645643  8.649599  8.667392  8.668112  8.698691
#>  [43]  8.667693  8.616479  8.599190  8.599200  8.596578  8.576172  8.685740
#>  [50] 11.632411 11.634097 11.616512 11.617393 11.688619 11.706103 11.706747
#>  [57] 12.725852 12.719576 12.720204 12.720813 12.721403 12.725274 12.769060
#>  [64] 12.783951 12.784458 12.788697 12.787167 12.784036 12.784219 12.783865
#>  [71] 12.785054 12.785449 12.785832 12.786203 12.786563 12.786911 12.787249
#>  [78] 12.787577 12.977633 12.976703 12.975798 12.974917 12.972227 12.972484
#>  [85] 12.971039 12.970249 12.969483 12.968740 12.968020 12.967323 12.966572
#>  [92] 12.968055 12.966867 12.966254 12.965663 12.962922 12.962360 12.961808
#>  [99] 12.959614 12.959089
#> 
#> $fold
#>  [1]  2  9  6  8  2  1  9  3 10  7  1  5  4  6  4  1  8  3  7 10  5
#> 
#> $lambda
#>   [1] 4.426334065 4.128013963 3.849799638 3.590336027 3.348359396 3.122691179
#>   [7] 2.912232245 2.715957538 2.532911090 2.362201360 2.202996894 2.054522277
#>  [13] 1.916054352 1.786918701 1.666486361 1.554170758 1.449424851 1.351738467
#>  [19] 1.260635818 1.175673183 1.096436746 1.022540579 0.953624766 0.889353648
#>  [25] 0.829414189 0.773514449 0.721382164 0.672763420 0.627421417 0.585135314
#>  [31] 0.545699152 0.508920856 0.474621293 0.442633406 0.412801396 0.384979964
#>  [37] 0.359033603 0.334835940 0.312269118 0.291223224 0.271595754 0.253291109
#>  [43] 0.236220136 0.220299689 0.205452228 0.191605435 0.178691870 0.166648636
#>  [49] 0.155417076 0.144942485 0.135173847 0.126063582 0.117567318 0.109643675
#>  [55] 0.102254058 0.095362477 0.088935365 0.082941419 0.077351445 0.072138216
#>  [61] 0.067276342 0.062742142 0.058513531 0.054569915 0.050892086 0.047462129
#>  [67] 0.044263341 0.041280140 0.038497996 0.035903360 0.033483594 0.031226912
#>  [73] 0.029122322 0.027159575 0.025329111 0.023622014 0.022029969 0.020545223
#>  [79] 0.019160544 0.017869187 0.016664864 0.015541708 0.014494249 0.013517385
#>  [85] 0.012606358 0.011756732 0.010964367 0.010225406 0.009536248 0.008893536
#>  [91] 0.008294142 0.007735144 0.007213822 0.006727634 0.006274214 0.005851353
#>  [97] 0.005456992 0.005089209 0.004746213 0.004426334
#> 
#> $fit
#> $beta
#>               4.4263     4.1280     3.8498     3.5903     3.3484    3.1227
#> (Intercept) 18.64762 20.2299323 21.8924229 23.4428672 24.8888168 26.237314
#> am           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> cyl          0.00000 -0.2407868 -0.4937745 -0.7297117 -0.9497475 -1.154954
#> disp         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> wt           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#>               2.9122    2.7160    2.5329    2.3622    2.2030    2.0545
#> (Intercept) 27.49493 28.667782 29.761590 30.781678 31.733017 32.620238
#> am           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.34633 -1.524807 -1.691256 -1.846487 -1.991256 -2.126268
#> disp         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.9161    1.7869    1.6665    1.5542   1.4494   1.3517   1.2606
#> (Intercept) 33.447663 34.219323 34.938976 35.610126 36.14733 36.14733 36.14733
#> am           0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#> carb         0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#> cyl         -2.252181 -2.369607 -2.479119 -2.581251 -2.66300 -2.66300 -2.66300
#> disp         0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#> drat         0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#> gear         0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#> hp           0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#> vs           0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#> wt           0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#>               1.1757      1.0964     1.0225     0.9536    0.8894    0.8294
#> (Intercept) 36.14733 36.10239178 35.8973819 36.8854074 37.258003 37.605492
#> am           0.00000  0.05664901  0.3150648  0.0000000  0.000000  0.000000
#> carb         0.00000  0.00000000  0.0000000  0.0000000  0.000000  0.000000
#> cyl         -2.66300 -2.65903457 -2.6409455 -2.3297819 -2.161272 -2.004005
#> disp         0.00000  0.00000000  0.0000000  0.0000000  0.000000  0.000000
#> drat         0.00000  0.00000000  0.0000000  0.0000000  0.000000  0.000000
#> gear         0.00000  0.00000000  0.0000000  0.0000000  0.000000  0.000000
#> hp           0.00000  0.00000000  0.0000000  0.0000000  0.000000  0.000000
#> qsec         0.00000  0.00000000  0.0000000  0.0000000  0.000000  0.000000
#> vs           0.00000  0.00000000  0.0000000  0.0000000  0.000000  0.000000
#> wt           0.00000  0.00000000  0.0000000 -0.8378917 -1.261430 -1.656641
#>                0.7735    0.7214    0.6728    0.6274    0.5851    0.5457
#> (Intercept) 37.929562 38.213210 38.213012 38.212912 38.212862 38.212836
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.857336 -1.728694 -1.728602 -1.728556 -1.728533 -1.728521
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.025217 -2.348323 -2.348439 -2.348498 -2.348527 -2.348542
#>                  0.5089      0.4746    0.4426     0.4128     0.3850     0.3590
#> (Intercept) 38.21107418 38.20857772 38.206145 38.2038731 38.2017547 34.7074225
#> am           0.00000000  0.00000000  0.000000  0.0000000  0.0000000  2.9819563
#> carb        -0.03850251 -0.08757154 -0.133437 -0.1762134 -0.2161069 -0.8832119
#> cyl         -1.70558791 -1.67624489 -1.648771 -1.6231471 -1.5992499 -1.4850911
#> disp         0.00000000  0.00000000  0.000000  0.0000000  0.0000000  0.0000000
#> drat         0.00000000  0.00000000  0.000000  0.0000000  0.0000000  0.0000000
#> gear         0.00000000  0.00000000  0.000000  0.0000000  0.0000000  0.0000000
#> hp           0.00000000  0.00000000  0.000000  0.0000000  0.0000000  0.0000000
#> qsec         0.00000000  0.00000000  0.000000  0.0000000  0.0000000  0.0000000
#> vs           0.00000000  0.00000000  0.000000  0.0000000  0.0000000  0.0000000
#> wt          -2.35601136 -2.36567743 -2.374771 -2.3832531 -2.3911637 -1.2811812
#>                 0.3348     0.3123     0.2912     0.2716     0.2533     0.2362
#> (Intercept) 34.7076984 34.7079188 34.7080877 34.7082131 34.7083038 34.7083678
#> am           2.9819384  2.9818765  2.9818022  2.9817312  2.9816702  2.9816211
#> carb        -0.8831302 -0.8830707 -0.8830283 -0.8829988 -0.8829787 -0.8829652
#> cyl         -1.4852546 -1.4853518 -1.4854074 -1.4854375 -1.4854525 -1.4854588
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -1.2810256 -1.2809543 -1.2809296 -1.2809291 -1.2809394 -1.2809535
#>                 0.2203     0.2055     0.1916     0.1787     0.1666     0.1554
#> (Intercept) 34.7084122 34.7084423 34.7084624 34.7084755 34.7084839 24.6597434
#> am           2.9815832  2.9815550  2.9815345  2.9815200  2.9815099  2.8768949
#> carb        -0.8829563 -0.8829505 -0.8829469 -0.8829447 -0.8829434 -1.1268011
#> cyl         -1.4854605 -1.4854599 -1.4854583 -1.4854564 -1.4854547 -0.6565936
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000 -0.0151492
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  1.3855091
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -1.2809675 -1.2809799 -1.2809900 -1.2809978 -1.2810038  0.0000000
#>                  0.1449      0.1352      0.1261      0.1176      0.1096
#> (Intercept) 24.65926302 24.65933816 24.65972238 24.66025608 24.66083923
#> am           2.87727256  2.87756891  2.87779882  2.87797549  2.87811011
#> carb        -1.12693083 -1.12700742 -1.12704867 -1.12706707 -1.12707125
#> cyl         -0.65656192 -0.65657967 -0.65662452 -0.65668191 -0.65674274
#> disp        -0.01514782 -0.01514663 -0.01514562 -0.01514479 -0.01514411
#> drat         1.38556888  1.38553534  1.38545153  1.38534500  1.38523272
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.1023      0.0954      0.0889      0.0829      0.0774
#> (Intercept) 24.66141205 24.66194146 24.66241162 24.66281748 24.66316041
#> am           2.87821193  2.87828842  2.87834549  2.87838783  2.87841903
#> carb        -1.12706706 -1.12705840 -1.12704774 -1.12703666 -1.12702604
#> cyl         -0.65680159 -0.65685550 -0.65690311 -0.65694405 -0.65697855
#> disp        -0.01514355 -0.01514311 -0.01514275 -0.01514248 -0.01514226
#> drat         1.38512462  1.38502601  1.38493926  1.38486491  1.38480245
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.0721      0.0673      0.0627      0.0585      0.0546
#> (Intercept) 24.66344532 24.66367879 24.66386790 24.66401958 24.66414019
#> am           2.87844188  2.87845852  2.87847055  2.87847918  2.87848534
#> carb        -1.12701639 -1.12700791 -1.12700065 -1.12699455 -1.12698951
#> cyl         -0.65700716 -0.65703057 -0.65704951 -0.65706469 -0.65707675
#> disp        -0.01514208 -0.01514195 -0.01514185 -0.01514177 -0.01514171
#> drat         1.38475080  1.38470864  1.38467462  1.38464741  1.38462583
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.0509      0.0475      0.0443      0.0413      0.0385
#> (Intercept) 24.66423534 24.66430989 24.66436792 24.66441283 24.66444739
#> am           2.87848969  2.87849273  2.87849484  2.87849628  2.87849726
#> carb        -1.12698539 -1.12698206 -1.12697940 -1.12697729 -1.12697563
#> cyl         -0.65708626 -0.65709370 -0.65709950 -0.65710398 -0.65710743
#> disp        -0.01514167 -0.01514163 -0.01514161 -0.01514159 -0.01514157
#> drat         1.38460885  1.38459558  1.38458527  1.38457731  1.38457120
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.0359      0.0335      0.0312      0.0291     0.0272
#> (Intercept) 24.66447384 24.66449399 24.66450925 28.11236452 28.1111638
#> am           2.87849790  2.87849831  2.87849857  2.62759463  2.6275145
#> carb        -1.12697433 -1.12697332 -1.12697254 -1.35833448 -1.3584860
#> cyl         -0.65711007 -0.65711207 -0.65711360 -0.75316715 -0.7530105
#> disp        -0.01514156 -0.01514155 -0.01514155 -0.01879662 -0.0188003
#> drat         1.38456653  1.38456298  1.38456029  1.19134519  1.1914492
#> gear         0.00000000  0.00000000  0.00000000  0.42376989  0.4240037
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> qsec         0.00000000  0.00000000  0.00000000 -0.23781874 -0.2378750
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> wt           0.00000000  0.00000000  0.00000000  0.66002314  0.6604281
#>                  0.0253      0.0236       0.0220       0.0205       0.0192
#> (Intercept) 28.11001874 28.10892867 28.938330993 28.938770289 28.939183578
#> am           2.62744099  2.62737359  2.674430262  2.674372140  2.674316694
#> carb        -1.35862732 -1.35875906 -1.449209638 -1.449369147 -1.449521558
#> cyl         -0.75286313 -0.75272466 -0.815449146 -0.815406413 -0.815365186
#> disp        -0.01880373 -0.01880692 -0.020650624 -0.020654589 -0.020658376
#> drat         1.19154562  1.19163498  1.153407569  1.153472875  1.153535294
#> gear         0.42422276  0.42442779  0.399223622  0.399354878  0.399480644
#> hp           0.00000000  0.00000000  0.003488348  0.003489948  0.003491475
#> qsec        -0.23792637 -0.23797332 -0.273012041 -0.273109708 -0.273202825
#> vs           0.00000000  0.00000000  0.000000000  0.000000000  0.000000000
#> wt           0.66080404  0.66115295  0.836745257  0.837203020  0.837640144
#>                  0.0179       0.0167       0.0155       0.0145       0.0135
#> (Intercept) 28.93957255 28.939938758 28.940283676 28.940608655 28.940914958
#> am           2.67426380  2.674213352  2.674165229  2.674119328  2.674075548
#> carb        -1.44966718 -1.449806288 -1.449939177 -1.450066109 -1.450187342
#> cyl         -0.81532543 -0.815287128 -0.815250236 -0.815214722 -0.815180553
#> disp        -0.02066199 -0.020665449 -0.020668748 -0.020671899 -0.020674907
#> drat         1.15359495  1.153651952  1.153706419  1.153758457  1.153808168
#> gear         0.39960112  0.399716513  0.399827005  0.399932786  0.400034040
#> hp           0.00349293  0.003494318  0.003495641  0.003496902  0.003498105
#> qsec        -0.27329160 -0.273376245 -0.273456943 -0.273533882 -0.273607236
#> vs           0.00000000  0.000000000  0.000000000  0.000000000  0.000000000
#> wt           0.83805753  0.838456052  0.838836532  0.839199769  0.839546526
#>                   0.0126       0.0118       0.0110       0.0102      0.0095
#> (Intercept) 28.941203761 28.941476160 28.941733173 28.941975753 28.94220479
#> am           2.674033791  2.673993965  2.673955982  2.673919756  2.67388521
#> carb        -1.450303125 -1.450413694 -1.450519277 -1.450620092 -1.45071635
#> cyl         -0.815147692 -0.815116103 -0.815085749 -0.815056592 -0.81502860
#> disp        -0.020677780 -0.020680522 -0.020683140 -0.020685640 -0.02068803
#> drat         1.153855653  1.153901008  1.153944325  1.153985692  1.15402519
#> gear         0.400130942  0.400223663  0.400312371  0.400397225  0.40047838
#> hp           0.003499252  0.003500345  0.003501388  0.003502382  0.00350333
#> qsec        -0.273677173 -0.273743853 -0.273807426 -0.273868038 -0.27392583
#> vs           0.000000000  0.000000000  0.000000000  0.000000000  0.00000000
#> wt           0.839877534  0.840193491  0.840495069  0.840782908  0.84105762
#>                   0.0089       0.0083       0.0077       0.0072      0.0067
#> (Intercept) 28.942421097 28.942625461 28.942818597 28.943001177 29.04975782
#> am           2.673852260  2.673820838  2.673790872  2.673762294  2.63626640
#> carb        -1.450808249 -1.450895984 -1.450979738 -1.451059688 -1.46739560
#> cyl         -0.815001724 -0.814975939 -0.814951205 -0.814927485 -0.84170793
#> disp        -0.020690303 -0.020692477 -0.020694551 -0.020696531 -0.02084690
#> drat         1.154062913  1.154098926  1.154133309  1.154166134  1.14769862
#> gear         0.400555992  0.400630200  0.400701145  0.400768965  0.42673479
#> hp           0.003504234  0.003505096  0.003505917  0.003506701  0.00386177
#> qsec        -0.273980924 -0.274033456 -0.274083541 -0.274131295 -0.27184666
#> vs           0.000000000  0.000000000  0.000000000  0.000000000 -0.11901101
#> wt           0.841319803  0.841570008  0.841808778  0.842036625  0.85424445
#>                   0.0063       0.0059       0.0055      0.0051       0.0047
#> (Intercept) 29.051820643 29.053792507 29.055676966 29.05747754 29.059197688
#> am           2.636116690  2.635971345  2.635830333  2.63569361  2.635561126
#> carb        -1.467585548 -1.467770998 -1.467951976 -1.46812851 -1.468300634
#> cyl         -0.841836041 -0.841956888 -0.842070842 -0.84217826 -0.842279487
#> disp        -0.020850982 -0.020854980 -0.020858894 -0.02086272 -0.020866469
#> drat         1.147691039  1.147686237  1.147683891  1.14768371  1.147685450
#> gear         0.426858244  0.426980110  0.427100324  0.42721882  0.427335540
#> hp           0.003865169  0.003868447  0.003871609  0.00387466  0.003877601
#> qsec        -0.271963742 -0.272077806 -0.272188857 -0.27229691 -0.272401989
#> vs          -0.119321350 -0.119618977 -0.119904378 -0.12017802 -0.120440363
#> wt           0.854730149  0.855205508  0.855670412  0.85612478  0.856568560
#>                   0.0044
#> (Intercept) 29.060840807
#> am           2.635432820
#> carb        -1.468468380
#> cyl         -0.842374855
#> disp        -0.020870128
#> drat         1.147688870
#> gear         0.427450413
#> hp           0.003880438
#> qsec        -0.272504125
#> vs          -0.120691842
#> wt           0.857001731
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   4   4  30  20  20  20  14   1   1   1   1   8   7   7   7   7
#>  [37]  26   1   1   1   1   1   1   1   1   1   1   1  66   1   1   1   1   1
#>  [55]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [73] 125   1   1   1  97   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1  50   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.426334065 4.128013963 3.849799638 3.590336027 3.348359396 3.122691179
#>   [7] 2.912232245 2.715957538 2.532911090 2.362201360 2.202996894 2.054522277
#>  [13] 1.916054352 1.786918701 1.666486361 1.554170758 1.449424851 1.351738467
#>  [19] 1.260635818 1.175673183 1.096436746 1.022540579 0.953624766 0.889353648
#>  [25] 0.829414189 0.773514449 0.721382164 0.672763420 0.627421417 0.585135314
#>  [31] 0.545699152 0.508920856 0.474621293 0.442633406 0.412801396 0.384979964
#>  [37] 0.359033603 0.334835940 0.312269118 0.291223224 0.271595754 0.253291109
#>  [43] 0.236220136 0.220299689 0.205452228 0.191605435 0.178691870 0.166648636
#>  [49] 0.155417076 0.144942485 0.135173847 0.126063582 0.117567318 0.109643675
#>  [55] 0.102254058 0.095362477 0.088935365 0.082941419 0.077351445 0.072138216
#>  [61] 0.067276342 0.062742142 0.058513531 0.054569915 0.050892086 0.047462129
#>  [67] 0.044263341 0.041280140 0.038497996 0.035903360 0.033483594 0.031226912
#>  [73] 0.029122322 0.027159575 0.025329111 0.023622014 0.022029969 0.020545223
#>  [79] 0.019160544 0.017869187 0.016664864 0.015541708 0.014494249 0.013517385
#>  [85] 0.012606358 0.011756732 0.010964367 0.010225406 0.009536248 0.008893536
#>  [91] 0.008294142 0.007735144 0.007213822 0.006727634 0.006274214 0.005851353
#>  [97] 0.005456992 0.005089209 0.004746213 0.004426334
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 22
#> 
#> $loss
#>   [1] 556.15238 486.18369 419.80783 364.49722 318.64753 280.87437 249.98393
#>   [8] 224.94744 204.87905 189.01661 176.70500 167.38165 160.56401 155.83869
#>  [15] 152.85198 151.30175 150.91987 150.91987 150.91987 150.91987 149.68600
#>  [22] 144.41465 121.98878 112.14472 105.85410 102.50615 101.57075 101.57075
#>  [29] 101.57075 101.57075 101.57075 100.15797  98.55272  97.25011  96.20751
#>  [36]  95.38502  77.86405  77.86405  77.86404  77.86404  77.86404  77.86404
#>  [43]  77.86404  77.86404  77.86404  77.86404  77.86404  77.86404  69.20304
#>  [50]  69.20304  69.20304  69.20304  69.20304  69.20304  69.20304  69.20304
#>  [57]  69.20304  69.20304  69.20304  69.20304  69.20304  69.20304  69.20304
#>  [64]  69.20304  69.20304  69.20304  69.20304  69.20304  69.20304  69.20304
#>  [71]  69.20304  69.20304  68.44773  68.44772  68.44771  68.44770  68.32814
#>  [78]  68.32813  68.32812  68.32811  68.32810  68.32809  68.32809  68.32808
#>  [85]  68.32807  68.32807  68.32806  68.32806  68.32805  68.32805  68.32805
#>  [92]  68.32804  68.32804  68.31925  68.31923  68.31922  68.31920  68.31919
#>  [99]  68.31918  68.31917
#> 
#> $linear.predictors
#>      4.4263   4.1280   3.8498   3.5903   3.3484   3.1227   2.9122   2.7160
#> 1  18.64762 18.78521 18.92978 19.06460 19.19033 19.30759 19.41695 19.51894
#> 2  18.64762 18.78521 18.92978 19.06460 19.19033 19.30759 19.41695 19.51894
#> 3  18.64762 18.78521 18.92978 19.06460 19.19033 19.30759 19.41695 19.51894
#> 4  18.64762 18.30364 17.94223 17.60517 17.29084 16.99769 16.72429 16.46932
#> 5  18.64762 19.26679 19.91732 20.52402 21.08983 21.61750 22.10961 22.56855
#> 6  18.64762 19.26679 19.91732 20.52402 21.08983 21.61750 22.10961 22.56855
#> 7  18.64762 18.78521 18.92978 19.06460 19.19033 19.30759 19.41695 19.51894
#> 8  18.64762 18.30364 17.94223 17.60517 17.29084 16.99769 16.72429 16.46932
#> 9  18.64762 18.30364 17.94223 17.60517 17.29084 16.99769 16.72429 16.46932
#> 10 18.64762 18.30364 17.94223 17.60517 17.29084 16.99769 16.72429 16.46932
#> 11 18.64762 18.30364 17.94223 17.60517 17.29084 16.99769 16.72429 16.46932
#> 12 18.64762 18.30364 17.94223 17.60517 17.29084 16.99769 16.72429 16.46932
#> 13 18.64762 18.30364 17.94223 17.60517 17.29084 16.99769 16.72429 16.46932
#> 14 18.64762 19.26679 19.91732 20.52402 21.08983 21.61750 22.10961 22.56855
#> 15 18.64762 19.26679 19.91732 20.52402 21.08983 21.61750 22.10961 22.56855
#> 16 18.64762 18.30364 17.94223 17.60517 17.29084 16.99769 16.72429 16.46932
#> 17 18.64762 18.30364 17.94223 17.60517 17.29084 16.99769 16.72429 16.46932
#> 18 18.64762 19.26679 19.91732 20.52402 21.08983 21.61750 22.10961 22.56855
#> 19 18.64762 18.30364 17.94223 17.60517 17.29084 16.99769 16.72429 16.46932
#> 20 18.64762 18.78521 18.92978 19.06460 19.19033 19.30759 19.41695 19.51894
#> 21 18.64762 18.30364 17.94223 17.60517 17.29084 16.99769 16.72429 16.46932
#>      2.5329   2.3622   2.2030   2.0545   1.9161   1.7869   1.6665   1.5542
#> 1  19.61405 19.70275 19.78548 19.86263 19.93458 20.00168 20.06426 20.12262
#> 2  19.61405 19.70275 19.78548 19.86263 19.93458 20.00168 20.06426 20.12262
#> 3  19.61405 19.70275 19.78548 19.86263 19.93458 20.00168 20.06426 20.12262
#> 4  16.23154 16.00978 15.80297 15.61009 15.43022 15.26247 15.10602 14.96012
#> 5  22.99656 23.39573 23.76799 24.11517 24.43894 24.74089 25.02250 25.28512
#> 6  22.99656 23.39573 23.76799 24.11517 24.43894 24.74089 25.02250 25.28512
#> 7  19.61405 19.70275 19.78548 19.86263 19.93458 20.00168 20.06426 20.12262
#> 8  16.23154 16.00978 15.80297 15.61009 15.43022 15.26247 15.10602 14.96012
#> 9  16.23154 16.00978 15.80297 15.61009 15.43022 15.26247 15.10602 14.96012
#> 10 16.23154 16.00978 15.80297 15.61009 15.43022 15.26247 15.10602 14.96012
#> 11 16.23154 16.00978 15.80297 15.61009 15.43022 15.26247 15.10602 14.96012
#> 12 16.23154 16.00978 15.80297 15.61009 15.43022 15.26247 15.10602 14.96012
#> 13 16.23154 16.00978 15.80297 15.61009 15.43022 15.26247 15.10602 14.96012
#> 14 22.99656 23.39573 23.76799 24.11517 24.43894 24.74089 25.02250 25.28512
#> 15 22.99656 23.39573 23.76799 24.11517 24.43894 24.74089 25.02250 25.28512
#> 16 16.23154 16.00978 15.80297 15.61009 15.43022 15.26247 15.10602 14.96012
#> 17 16.23154 16.00978 15.80297 15.61009 15.43022 15.26247 15.10602 14.96012
#> 18 22.99656 23.39573 23.76799 24.11517 24.43894 24.74089 25.02250 25.28512
#> 19 16.23154 16.00978 15.80297 15.61009 15.43022 15.26247 15.10602 14.96012
#> 20 19.61405 19.70275 19.78548 19.86263 19.93458 20.00168 20.06426 20.12262
#> 21 16.23154 16.00978 15.80297 15.61009 15.43022 15.26247 15.10602 14.96012
#>      1.4494   1.3517   1.2606   1.1757   1.0964   1.0225   0.9536   0.8894
#> 1  20.16933 20.16933 20.16933 20.16933 20.20483 20.36677 20.71144 20.98542
#> 2  20.16933 20.16933 20.16933 20.16933 20.20483 20.36677 20.49778 20.66376
#> 3  20.16933 20.16933 20.16933 20.16933 20.14818 20.05171 20.00761 19.92582
#> 4  14.84333 14.84333 14.84333 14.84333 14.83012 14.76982 15.25588 15.46452
#> 5  25.49533 25.49533 25.49533 25.49533 25.46625 25.33360 24.89341 24.58895
#> 6  25.49533 25.49533 25.49533 25.49533 25.46625 25.33360 24.92692 24.63941
#> 7  20.16933 20.16933 20.16933 20.16933 20.14818 20.05171 20.02437 19.95105
#> 8  14.84333 14.84333 14.84333 14.84333 14.83012 14.76982 14.83693 14.83380
#> 9  14.84333 14.84333 14.84333 14.84333 14.83012 14.76982 15.12182 15.26269
#> 10 14.84333 14.84333 14.84333 14.84333 14.83012 14.76982 15.07992 15.19962
#> 11 14.84333 14.84333 14.84333 14.84333 14.83012 14.76982 13.84822 13.34532
#> 12 14.84333 14.84333 14.84333 14.84333 14.83012 14.76982 13.70243 13.12583
#> 13 14.84333 14.84333 14.84333 14.84333 14.83012 14.76982 13.76862 13.22548
#> 14 25.49533 25.49533 25.49533 25.49533 25.52290 25.64866 25.72292 25.83777
#> 15 25.49533 25.49533 25.49533 25.49533 25.46625 25.33360 25.50088 25.50349
#> 16 14.84333 14.84333 14.84333 14.84333 14.83012 14.76982 15.29777 15.52759
#> 17 14.84333 14.84333 14.84333 14.84333 14.83012 14.76982 15.02546 15.11763
#> 18 25.49533 25.49533 25.49533 25.49533 25.52290 25.64866 25.94496 26.17205
#> 19 14.84333 14.84333 14.84333 14.84333 14.88676 15.08488 15.59104 15.96909
#> 20 20.16933 20.16933 20.16933 20.16933 20.20483 20.36677 20.58576 20.79621
#> 21 14.84333 14.84333 14.84333 14.84333 14.88676 15.08488 15.25588 15.46452
#>      0.8294   0.7735   0.7214   0.6728   0.6274   0.5851   0.5457   0.5089
#> 1  21.24106 21.47948 21.68844 21.68849 21.68851 21.68852 21.68853 21.65079
#> 2  20.81862 20.96305 21.08962 21.08963 21.08964 21.08965 21.08965 21.05000
#> 3  19.84948 19.77829 19.71585 19.71580 19.71577 19.71576 19.71575 19.78724
#> 4  15.65924 15.84085 16.00014 16.00026 16.00033 16.00036 16.00037 16.00140
#> 5  24.30479 24.03977 23.80728 23.80708 23.80698 23.80693 23.80690 23.79604
#> 6  24.37105 24.12078 23.90122 23.90102 23.90092 23.90087 23.90085 23.89028
#> 7  19.88262 19.81880 19.76281 19.76277 19.76274 19.76273 19.76273 19.71886
#> 8  14.83092 14.82824 14.82598 14.82605 14.82608 14.82609 14.82610 14.86190
#> 9  15.39418 15.51681 15.62441 15.62451 15.62457 15.62459 15.62461 15.66294
#> 10 15.31135 15.41555 15.50699 15.50709 15.50714 15.50717 15.50718 15.54514
#> 11 12.87609 12.43848 12.05496 12.05489 12.05485 12.05483 12.05482 12.04330
#> 12 12.58783 12.08609 11.64635 11.64626 11.64621 11.64619 11.64618 11.63336
#> 13 12.71871 12.24609 11.83187 11.83179 11.83174 11.83172 11.83171 11.81948
#> 14 25.94486 26.04474 26.13212 26.13204 26.13199 26.13197 26.13196 26.16700
#> 15 25.50585 25.50806 25.50982 25.50970 25.50964 25.50961 25.50960 25.54265
#> 16 15.74208 15.94211 16.11756 16.11769 16.11775 16.11778 16.11780 16.19621
#> 17 15.20367 15.28391 15.35435 15.35444 15.35449 15.35451 15.35452 15.43050
#> 18 26.38387 26.58142 26.75443 26.75437 26.75434 26.75433 26.75432 26.79134
#> 19 16.32190 16.65093 16.93947 16.93964 16.93973 16.93977 16.93979 16.94380
#> 20 20.99257 21.17569 21.33619 21.33622 21.33624 21.33624 21.33625 21.22038
#> 21 15.65924 15.84085 16.00014 16.00026 16.00033 16.00036 16.00037 15.84739
#>      0.4746   0.4426   0.4128   0.3850   0.3590   0.3348   0.3123   0.2912
#> 1  21.60275 21.55787 21.51601 21.47698 21.88929 21.88930 21.88930 21.88930
#> 2  20.99950 20.95230 20.90828 20.86723 21.56259 21.56264 21.56266 21.56266
#> 3  19.87829 19.96337 20.04272 20.11672 20.48078 20.48069 20.48064 20.48060
#> 4  16.00286 16.00429 16.00563 16.00687 14.72003 14.71988 14.71981 14.71980
#> 5  23.78194 23.76867 23.75628 23.74473 22.91367 22.91395 22.91413 22.91424
#> 6  23.87657 23.86366 23.85161 23.84038 22.96491 22.96519 22.96536 22.96547
#> 7  19.66289 19.61056 19.56175 19.51622 17.85676 17.85692 17.85704 17.85713
#> 8  14.90760 14.95035 14.99022 15.02740 14.96265 14.96250 14.96241 14.96236
#> 9  15.71193 15.75777 15.80052 15.84039 15.39825 15.39805 15.39793 15.39788
#> 10 15.59364 15.63903 15.68136 15.72084 15.33419 15.33399 15.33389 15.33383
#> 11 12.02853 12.01468 12.00176 11.98972 12.56764 12.56776 12.56781 12.56783
#> 12 11.61690 11.60147 11.58708 11.57366 12.34472 12.34486 12.34493 12.34495
#> 13 11.80379 11.78908 11.77535 11.76256 12.44593 12.44606 12.44612 12.44615
#> 14 26.21154 26.25313 26.29191 26.32809 28.04720 28.04723 28.04722 28.04719
#> 15 25.58463 25.62381 25.66035 25.69443 24.72573 24.72582 24.72589 24.72594
#> 16 16.29629 16.38991 16.47722 16.55865 16.55051 16.55019 16.55000 16.54990
#> 17 15.52745 15.61811 15.70266 15.78152 16.13413 16.13386 16.13369 16.13360
#> 18 26.83844 26.88244 26.92348 26.96175 28.38672 28.38670 28.38667 28.38663
#> 19 16.94913 16.95420 16.95893 16.96334 18.21446 18.21423 18.21407 18.21397
#> 20 21.07275 20.93478 20.80610 20.68609 19.93069 19.93089 19.93102 19.93110
#> 21 15.65258 15.47055 15.30078 15.14245 14.16914 14.16930 14.16941 14.16949
#>      0.2716   0.2533   0.2362   0.2203   0.2055   0.1916   0.1787   0.1666
#> 1  21.88929 21.88928 21.88928 21.88927 21.88927 21.88927 21.88926 21.88926
#> 2  21.56265 21.56264 21.56263 21.56263 21.56262 21.56261 21.56261 21.56261
#> 3  20.48057 20.48056 20.48055 20.48055 20.48054 20.48054 20.48054 20.48054
#> 4  14.71980 14.71982 14.71983 14.71985 14.71986 14.71987 14.71988 14.71989
#> 5  22.91430 22.91434 22.91436 22.91437 22.91438 22.91438 22.91438 22.91438
#> 6  22.96554 22.96558 22.96560 22.96561 22.96562 22.96562 22.96562 22.96562
#> 7  17.85720 17.85724 17.85727 17.85730 17.85731 17.85732 17.85733 17.85733
#> 8  14.96234 14.96232 14.96232 14.96232 14.96232 14.96233 14.96233 14.96233
#> 9  15.39785 15.39784 15.39785 15.39785 15.39786 15.39786 15.39787 15.39787
#> 10 15.33380 15.33380 15.33380 15.33380 15.33381 15.33381 15.33382 15.33382
#> 11 12.56784 12.56784 12.56783 12.56782 12.56782 12.56781 12.56781 12.56780
#> 12 12.34496 12.34495 12.34494 12.34494 12.34493 12.34492 12.34491 12.34491
#> 13 12.44615 12.44615 12.44614 12.44613 12.44612 12.44612 12.44611 12.44611
#> 14 28.04715 28.04712 28.04709 28.04707 28.04705 28.04704 28.04703 28.04702
#> 15 24.72597 24.72600 24.72602 24.72603 24.72604 24.72604 24.72605 24.72605
#> 16 16.54984 16.54982 16.54981 16.54981 16.54981 16.54982 16.54982 16.54983
#> 17 16.13354 16.13351 16.13350 16.13350 16.13349 16.13350 16.13350 16.13350
#> 18 28.38660 28.38657 28.38654 28.38652 28.38651 28.38650 28.38649 28.38649
#> 19 18.21390 18.21386 18.21383 18.21382 18.21381 18.21380 18.21380 18.21380
#> 20 19.93115 19.93118 19.93120 19.93121 19.93122 19.93122 19.93122 19.93122
#> 21 14.16954 14.16957 14.16959 14.16961 14.16962 14.16962 14.16962 14.16963
#>      0.1554   0.1449   0.1352   0.1261   0.1176   0.1096   0.1023   0.0954
#> 1  22.06949 22.06951 22.06953 22.06954 22.06955 22.06956 22.06956 22.06957
#> 2  22.06949 22.06951 22.06953 22.06954 22.06955 22.06956 22.06956 22.06957
#> 3  20.00882 20.00887 20.00894 20.00901 20.00907 20.00913 20.00918 20.00922
#> 4  13.89356 13.89351 13.89345 13.89341 13.89337 13.89333 13.89330 13.89328
#> 5  22.66991 22.66972 22.66962 22.66958 22.66958 22.66959 22.66962 22.66965
#> 6  23.07795 23.07777 23.07766 23.07759 23.07756 23.07755 23.07755 23.07755
#> 7  19.10517 19.10482 19.10455 19.10434 19.10418 19.10406 19.10396 19.10389
#> 8  16.10195 16.10190 16.10183 16.10175 16.10168 16.10160 16.10154 16.10148
#> 9  16.10195 16.10190 16.10183 16.10175 16.10168 16.10160 16.10154 16.10148
#> 10 16.10195 16.10190 16.10183 16.10175 16.10168 16.10160 16.10154 16.10148
#> 11 11.80891 11.80899 11.80908 11.80917 11.80925 11.80933 11.80939 11.80944
#> 12 12.08768 12.08776 12.08783 12.08790 12.08796 12.08802 12.08807 12.08811
#> 13 12.70934 12.70939 12.70943 12.70947 12.70949 12.70951 12.70952 12.70953
#> 14 28.24410 28.24434 28.24453 28.24466 28.24475 28.24482 28.24486 28.24489
#> 15 24.21353 24.21344 24.21338 24.21336 24.21335 24.21335 24.21336 24.21337
#> 16 16.15995 16.16007 16.16014 16.16017 16.16018 16.16017 16.16016 16.16014
#> 17 15.36108 15.36133 15.36148 15.36157 15.36161 15.36163 15.36163 15.36162
#> 18 28.23955 28.23980 28.23998 28.24011 28.24021 28.24027 28.24032 28.24035
#> 19 18.30616 18.30653 18.30673 18.30682 18.30684 18.30682 18.30678 18.30673
#> 20 19.65518 19.65491 19.65476 19.65470 19.65469 19.65471 19.65475 19.65479
#> 21 13.61427 13.61401 13.61387 13.61380 13.61378 13.61379 13.61381 13.61384
#>      0.0889   0.0829   0.0774   0.0721   0.0673   0.0627   0.0585   0.0546
#> 1  22.06957 22.06957 22.06957 22.06957 22.06957 22.06957 22.06957 22.06957
#> 2  22.06957 22.06957 22.06957 22.06957 22.06957 22.06957 22.06957 22.06957
#> 3  20.00926 20.00929 20.00931 20.00933 20.00934 20.00936 20.00937 20.00937
#> 4  13.89326 13.89324 13.89323 13.89322 13.89321 13.89321 13.89320 13.89320
#> 5  22.66969 22.66972 22.66975 22.66977 22.66979 22.66981 22.66982 22.66983
#> 6  23.07757 23.07758 23.07759 23.07760 23.07761 23.07762 23.07763 23.07763
#> 7  19.10384 19.10380 19.10377 19.10375 19.10373 19.10372 19.10371 19.10370
#> 8  16.10144 16.10140 16.10136 16.10134 16.10132 16.10130 16.10129 16.10127
#> 9  16.10144 16.10140 16.10136 16.10134 16.10132 16.10130 16.10129 16.10127
#> 10 16.10144 16.10140 16.10136 16.10134 16.10132 16.10130 16.10129 16.10127
#> 11 11.80949 11.80952 11.80955 11.80958 11.80960 11.80961 11.80962 11.80963
#> 12 12.08815 12.08817 12.08820 12.08822 12.08823 12.08824 12.08825 12.08826
#> 13 12.70954 12.70954 12.70955 12.70955 12.70955 12.70955 12.70956 12.70956
#> 14 28.24491 28.24493 28.24494 28.24494 28.24495 28.24495 28.24495 28.24495
#> 15 24.21338 24.21339 24.21340 24.21341 24.21342 24.21343 24.21343 24.21344
#> 16 16.16013 16.16011 16.16010 16.16008 16.16007 16.16006 16.16006 16.16005
#> 17 15.36160 15.36159 15.36157 15.36155 15.36154 15.36153 15.36152 15.36151
#> 18 28.24037 28.24039 28.24039 28.24040 28.24040 28.24041 28.24041 28.24041
#> 19 18.30668 18.30663 18.30658 18.30654 18.30651 18.30648 18.30645 18.30643
#> 20 19.65483 19.65487 19.65491 19.65494 19.65497 19.65499 19.65501 19.65503
#> 21 13.61387 13.61390 13.61392 13.61395 13.61397 13.61399 13.61400 13.61402
#>      0.0509   0.0475   0.0443   0.0413   0.0385   0.0359   0.0335   0.0312
#> 1  22.06957 22.06957 22.06957 22.06957 22.06957 22.06957 22.06957 22.06957
#> 2  22.06957 22.06957 22.06957 22.06957 22.06957 22.06957 22.06957 22.06957
#> 3  20.00938 20.00938 20.00939 20.00939 20.00939 20.00939 20.00939 20.00939
#> 4  13.89320 13.89320 13.89319 13.89319 13.89319 13.89319 13.89319 13.89319
#> 5  22.66984 22.66985 22.66986 22.66986 22.66987 22.66987 22.66987 22.66987
#> 6  23.07764 23.07764 23.07765 23.07765 23.07765 23.07765 23.07766 23.07766
#> 7  19.10370 19.10370 19.10369 19.10369 19.10369 19.10369 19.10369 19.10369
#> 8  16.10127 16.10126 16.10126 16.10125 16.10125 16.10125 16.10125 16.10124
#> 9  16.10127 16.10126 16.10126 16.10125 16.10125 16.10125 16.10125 16.10124
#> 10 16.10127 16.10126 16.10126 16.10125 16.10125 16.10125 16.10125 16.10124
#> 11 11.80964 11.80965 11.80965 11.80965 11.80966 11.80966 11.80966 11.80966
#> 12 12.08826 12.08827 12.08827 12.08827 12.08828 12.08828 12.08828 12.08828
#> 13 12.70956 12.70956 12.70956 12.70956 12.70956 12.70956 12.70956 12.70956
#> 14 28.24495 28.24495 28.24495 28.24495 28.24495 28.24495 28.24495 28.24495
#> 15 24.21344 24.21345 24.21345 24.21345 24.21345 24.21345 24.21345 24.21346
#> 16 16.16005 16.16004 16.16004 16.16004 16.16003 16.16003 16.16003 16.16003
#> 17 15.36150 15.36150 15.36149 15.36149 15.36149 15.36148 15.36148 15.36148
#> 18 28.24041 28.24041 28.24041 28.24041 28.24041 28.24041 28.24040 28.24040
#> 19 18.30642 18.30641 18.30640 18.30639 18.30638 18.30638 18.30637 18.30637
#> 20 19.65504 19.65505 19.65506 19.65506 19.65507 19.65507 19.65507 19.65507
#> 21 13.61403 13.61403 13.61404 13.61404 13.61405 13.61405 13.61405 13.61406
#>      0.0291   0.0272   0.0253   0.0236   0.0220   0.0205   0.0192   0.0179
#> 1  21.93625 21.93619 21.93613 21.93608 21.79652 21.79644 21.79635 21.79627
#> 2  21.97138 21.97139 21.97140 21.97141 21.85701 21.85698 21.85696 21.85694
#> 3  20.04020 20.04021 20.04022 20.04023 20.07222 20.07222 20.07221 20.07221
#> 4  13.57167 13.57138 13.57111 13.57086 13.60110 13.60090 13.60070 13.60052
#> 5  23.06581 23.06587 23.06594 23.06599 23.02689 23.02697 23.02705 23.02712
#> 6  22.73464 22.73457 22.73452 22.73446 22.70392 22.70380 22.70367 22.70356
#> 7  19.29326 19.29348 19.29369 19.29388 19.21736 19.21751 19.21767 19.21781
#> 8  16.30490 16.30518 16.30543 16.30567 16.39335 16.39360 16.39384 16.39408
#> 9  16.03293 16.03306 16.03317 16.03328 16.05425 16.05433 16.05441 16.05448
#> 10 15.97081 15.97093 15.97104 15.97115 15.98688 15.98695 15.98701 15.98707
#> 11 11.73278 11.73261 11.73245 11.73230 11.64723 11.64706 11.64690 11.64675
#> 12 12.19463 12.19459 12.19455 12.19451 12.19993 12.19993 12.19992 12.19992
#> 13 12.88755 12.88760 12.88764 12.88768 12.97366 12.97377 12.97388 12.97399
#> 14 28.26715 28.26721 28.26727 28.26732 28.33487 28.33496 28.33504 28.33511
#> 15 24.03138 24.03117 24.03097 24.03079 24.15044 24.15038 24.15032 24.15026
#> 16 16.26374 16.26378 16.26382 16.26386 16.19338 16.19336 16.19334 16.19332
#> 17 15.27534 15.27524 15.27514 15.27505 15.17913 15.17897 15.17883 15.17869
#> 18 28.22216 28.22215 28.22213 28.22212 28.26256 28.26257 28.26259 28.26261
#> 19 18.47390 18.47405 18.47419 18.47432 18.52219 18.52227 18.52235 18.52243
#> 20 19.91903 19.91904 19.91905 19.91906 19.89848 19.89851 19.89854 19.89857
#> 21 13.41050 13.41032 13.41014 13.40998 13.52864 13.52852 13.52841 13.52831
#>      0.0167   0.0155   0.0145   0.0135   0.0126   0.0118   0.0110   0.0102
#> 1  21.79620 21.79613 21.79606 21.79599 21.79593 21.79586 21.79581 21.79575
#> 2  21.85691 21.85689 21.85687 21.85685 21.85684 21.85682 21.85680 21.85679
#> 3  20.07221 20.07220 20.07220 20.07220 20.07219 20.07219 20.07219 20.07219
#> 4  13.60034 13.60017 13.60001 13.59986 13.59971 13.59957 13.59944 13.59931
#> 5  23.02720 23.02726 23.02733 23.02739 23.02745 23.02750 23.02756 23.02761
#> 6  22.70344 22.70334 22.70324 22.70314 22.70305 22.70296 22.70288 22.70280
#> 7  19.21795 19.21808 19.21821 19.21833 19.21844 19.21855 19.21866 19.21876
#> 8  16.39430 16.39451 16.39472 16.39491 16.39509 16.39527 16.39544 16.39560
#> 9  16.05455 16.05462 16.05468 16.05474 16.05480 16.05486 16.05491 16.05496
#> 10 15.98712 15.98718 15.98723 15.98728 15.98732 15.98737 15.98741 15.98745
#> 11 11.64660 11.64646 11.64633 11.64620 11.64608 11.64596 11.64585 11.64574
#> 12 12.19991 12.19991 12.19991 12.19990 12.19990 12.19989 12.19989 12.19989
#> 13 12.97409 12.97419 12.97428 12.97437 12.97445 12.97453 12.97461 12.97468
#> 14 28.33519 28.33525 28.33532 28.33538 28.33544 28.33550 28.33556 28.33561
#> 15 24.15021 24.15016 24.15011 24.15006 24.15001 24.14997 24.14993 24.14988
#> 16 16.19330 16.19328 16.19327 16.19325 16.19324 16.19323 16.19321 16.19320
#> 17 15.17855 15.17842 15.17830 15.17819 15.17807 15.17797 15.17787 15.17777
#> 18 28.26262 28.26263 28.26265 28.26266 28.26267 28.26268 28.26269 28.26270
#> 19 18.52250 18.52257 18.52264 18.52270 18.52276 18.52282 18.52288 18.52293
#> 20 19.89860 19.89863 19.89865 19.89868 19.89870 19.89872 19.89874 19.89876
#> 21 13.52821 13.52811 13.52802 13.52793 13.52785 13.52777 13.52769 13.52762
#>      0.0095   0.0089   0.0083   0.0077   0.0072   0.0067   0.0063   0.0059
#> 1  21.79570 21.79565 21.79560 21.79555 21.79551 21.80197 21.80189 21.80181
#> 2  21.85677 21.85676 21.85674 21.85673 21.85672 21.86757 21.86754 21.86752
#> 3  20.07218 20.07218 20.07218 20.07218 20.07218 20.03484 20.03473 20.03463
#> 4  13.59918 13.59907 13.59895 13.59885 13.59874 13.59568 13.59555 13.59541
#> 5  23.02765 23.02770 23.02774 23.02779 23.02783 23.04037 23.04051 23.04065
#> 6  22.70272 22.70265 22.70258 22.70252 22.70245 22.73225 22.73217 22.73209
#> 7  19.21885 19.21894 19.21903 19.21911 19.21919 19.16170 19.16165 19.16161
#> 8  16.39576 16.39590 16.39604 16.39618 16.39631 16.40973 16.40997 16.41021
#> 9  16.05501 16.05506 16.05510 16.05515 16.05519 16.06492 16.06497 16.06503
#> 10 15.98749 15.98753 15.98757 15.98760 15.98764 15.99889 15.99892 15.99895
#> 11 11.64564 11.64554 11.64545 11.64536 11.64528 11.63838 11.63822 11.63807
#> 12 12.19988 12.19988 12.19988 12.19987 12.19987 12.19963 12.19966 12.19969
#> 13 12.97475 12.97481 12.97488 12.97494 12.97499 12.97972 12.97989 12.98005
#> 14 28.33566 28.33571 28.33575 28.33580 28.33584 28.32304 28.32310 28.32315
#> 15 24.14984 24.14981 24.14977 24.14974 24.14970 24.16015 24.16023 24.16032
#> 16 16.19319 16.19318 16.19317 16.19316 16.19315 16.19999 16.19995 16.19990
#> 17 15.17768 15.17759 15.17751 15.17743 15.17735 15.18305 15.18289 15.18273
#> 18 28.26271 28.26272 28.26273 28.26274 28.26274 28.24537 28.24536 28.24535
#> 19 18.52298 18.52303 18.52308 18.52313 18.52317 18.52816 18.52818 18.52821
#> 20 19.89878 19.89880 19.89882 19.89883 19.89885 19.92539 19.92552 19.92564
#> 21 13.52755 13.52748 13.52742 13.52736 13.52730 13.50919 13.50908 13.50898
#>      0.0055   0.0051   0.0047   0.0044
#> 1  21.80172 21.80165 21.80157 21.80149
#> 2  21.86749 21.86747 21.86745 21.86743
#> 3  20.03453 20.03444 20.03435 20.03426
#> 4  13.59528 13.59515 13.59502 13.59489
#> 5  23.04079 23.04092 23.04105 23.04118
#> 6  22.73201 22.73194 22.73187 22.73179
#> 7  19.16157 19.16154 19.16151 19.16149
#> 8  16.41044 16.41067 16.41090 16.41112
#> 9  16.06508 16.06513 16.06518 16.06523
#> 10 15.99899 15.99902 15.99905 15.99908
#> 11 11.63791 11.63776 11.63762 11.63747
#> 12 12.19971 12.19973 12.19976 12.19978
#> 13 12.98021 12.98036 12.98051 12.98065
#> 14 28.32321 28.32326 28.32331 28.32336
#> 15 24.16039 24.16046 24.16052 24.16058
#> 16 16.19986 16.19982 16.19978 16.19974
#> 17 15.18258 15.18243 15.18228 15.18214
#> 18 28.24535 28.24534 28.24533 28.24533
#> 19 18.52823 18.52825 18.52828 18.52831
#> 20 19.92576 19.92587 19.92598 19.92609
#> 21 13.50888 13.50877 13.50867 13.50857
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 18.1 14.3 24.4 22.8 19.2 16.4 17.3 15.2 10.4 10.4 14.7 32.4 21.5
#> [16] 15.5 19.2 27.3 15.8 19.7 15.0
#> 
#> $X
#>               am       carb        cyl       disp       drat       gear
#>  [1,]  1.4142136  0.4740547 -0.3464102 -0.7737596  0.9425590  0.5275044
#>  [2,]  1.4142136  0.4740547 -0.3464102 -0.7737596  0.9425590  0.5275044
#>  [3,] -0.7071068 -1.2827362 -0.3464102 -0.2436277 -1.5235461 -0.8571946
#>  [4,] -0.7071068  0.4740547  0.8660254  0.8574156 -0.5500836 -0.8571946
#>  [5,] -0.7071068 -0.6971392 -1.5588457 -0.8822327  0.4882764  0.5275044
#>  [6,] -0.7071068 -0.6971392 -1.5588457 -0.9303524  0.9858240  0.5275044
#>  [7,] -0.7071068  0.4740547 -0.3464102 -0.7117749  0.9858240  0.5275044
#>  [8,] -0.7071068 -0.1115423  0.8660254  0.1706908 -0.8529386 -0.8571946
#>  [9,] -0.7071068 -0.1115423  0.8660254  0.1706908 -0.8529386 -0.8571946
#> [10,] -0.7071068 -0.1115423  0.8660254  0.1706908 -0.8529386 -0.8571946
#> [11,] -0.7071068  0.4740547  0.8660254  1.7708736 -1.1557936 -0.8571946
#> [12,] -0.7071068  0.4740547  0.8660254  1.6730031 -1.0043661 -0.8571946
#> [13,] -0.7071068  0.4740547  0.8660254  1.5098856 -0.5068186 -0.8571946
#> [14,]  1.4142136 -1.2827362 -1.5588457 -1.4368323  1.3319440  0.5275044
#> [15,] -0.7071068 -1.2827362 -1.5588457 -1.0991790  0.5099089 -0.8571946
#> [16,] -0.7071068 -0.6971392  0.8660254  0.5148688 -1.5235461 -0.8571946
#> [17,] -0.7071068 -0.6971392  0.8660254  1.1836506 -0.8313061 -0.8571946
#> [18,]  1.4142136 -1.2827362 -1.5588457 -1.4343855  1.3319440  0.5275044
#> [19,]  1.4142136  0.4740547  0.8660254  0.7840127  1.6347990  1.9122034
#> [20,]  1.4142136  1.6452486 -0.3464102 -0.8960977  0.3368489  1.9122034
#> [21,]  1.4142136  2.8164425  0.8660254  0.3762189  0.1637889  1.9122034
#>               hp         qsec         vs          wt
#>  [1,] -0.7117255 -0.686511892 -0.7071068 -0.94443976
#>  [2,] -0.7117255 -0.400221826 -0.7071068 -0.66896316
#>  [3,] -0.7823599  1.235721406  1.4142136 -0.03698743
#>  [4,]  1.1954028 -1.003475893 -0.7071068  0.08184562
#>  [5,] -1.3898156  1.123250309  1.4142136 -0.32866854
#>  [6,] -0.9236286  2.605823863  1.4142136 -0.37188055
#>  [7,] -0.5280761  0.254155466  1.4142136 -0.05859344
#>  [8,]  0.2771559 -0.205953568 -0.7071068  0.62199582
#>  [9,]  0.2771559 -0.103707116 -0.7071068  0.25469368
#> [10,]  0.2771559  0.100785788 -0.7071068  0.30870870
#> [11,]  0.6303278  0.090561143 -0.7071068  1.89675029
#> [12,]  0.7715965  0.008763982 -0.7071068  2.08472256
#> [13,]  0.9834997 -0.195728922 -0.7071068  1.99937882
#> [14,] -1.3333081  0.852297211  1.4142136 -1.39816593
#> [15,] -0.8953749  1.128362631  1.4142136 -1.11188633
#> [16,] -0.1466504 -0.476906665 -0.7071068  0.02783060
#> [17,]  0.2065215 -0.384884859 -0.7071068  0.37892823
#> [18,] -1.3333081  0.560894822  1.4142136 -1.68444554
#> [19,]  1.4638135 -1.688527122 -0.7071068 -0.35027454
#> [20,]  0.2065215 -1.177294862 -0.7071068 -0.78239470
#> [21,]  2.4668217 -1.637403896 -0.7071068  0.08184562
#> attr(,"center")
#>  [1]   0.3333333   3.1904762   6.5714286 254.8714286   3.4642857   3.6190476
#>  [7] 160.3809524  17.8028571   0.3333333   3.4942381
#> attr(,"scale")
#>  [1]   0.4714045   1.7076592   1.6495722 122.6109921   0.4622674   0.7221786
#>  [7]  70.7870557   1.9560581   0.4714045   0.9256685
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 30
#> 
#> $lambda.min
#> [1] 0.5851353
#> 
#> $null.dev
#> [1] 26.48345
#> 
#> $Bias
#> [1] 5.970842
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 9.113508