Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 45.92754 43.56608 40.59431 37.30213 34.10784 31.57347 29.59664 28.08939
#>   [9] 26.97627 26.19260 25.68306 25.40033 25.30405 25.35987 25.53856 25.43077
#>  [17] 25.17762 24.79829 24.48031 24.30949 24.03325 23.63015 23.23681 22.88973
#>  [25] 22.59845 21.86367 21.64199 21.43847 21.43185 20.83232 16.73284 15.63984
#>  [33] 15.15702 14.88028 14.67692 14.51209 14.40057 14.35570 14.38659 12.35786
#>  [41] 12.33864 12.16087 12.16043 12.91714 12.91784 19.15143 19.60319 20.02760
#>  [49] 28.88949 28.89207 31.48239 38.83045 43.54597 43.55018 44.10555 44.17706
#>  [57] 44.40423 44.41090 53.35024 54.50153 54.50545 54.50912 58.47256 58.51280
#>  [65] 59.09924 59.42763 60.29481 60.41187 60.30699 60.32437 60.53320 61.04520
#>  [73] 61.04637 62.96999 63.91340 63.91493 63.93859 63.94251 63.94633 63.95006
#>  [81] 63.95370 64.34121 64.37435 64.57586 64.58027 64.58458 64.58881 64.59296
#>  [89] 64.59708 64.60681 64.61082 64.68966 64.71311 64.71719 64.72119 64.72510
#>  [97] 64.72892 64.73267 64.73634 64.73993
#> 
#> $cvse
#>   [1] 13.604479 13.374877 13.006619 12.124028 10.973920  9.930410  8.988297
#>   [8]  8.146567  7.407704  6.776884  6.260788  5.865762  5.595392  5.448056
#>  [15]  5.415562  5.363165  5.347100  5.235101  5.142124  5.149004  5.170786
#>  [22]  5.153374  5.128372  5.108338  5.088566  5.080241  5.061850  5.038075
#>  [29]  5.009351  4.921744  4.287455  4.213503  4.113048  4.008172  3.951473
#>  [36]  3.916114  3.920010  3.932004  3.934057  3.317997  3.117453  3.043689
#>  [43]  3.043526  3.193045  3.193150  7.213569  7.268837  7.292570 10.156939
#>  [50] 10.158333 11.259114 16.455238 18.115088 18.117645 18.067121 18.065621
#>  [57] 18.047669 18.052269 22.080608 22.552342 22.554433 22.556473 23.389704
#>  [64] 23.387469 23.338140 23.322845 23.277867 23.270038 23.282491 23.279919
#>  [71] 23.259098 23.379711 23.381419 23.258483 23.229711 23.231328 23.234881
#>  [78] 23.236222 23.237526 23.238795 23.240029 23.211607 23.209008 23.291809
#>  [85] 23.292934 23.294030 23.295098 23.296140 23.297148 23.297322 23.298278
#>  [92] 23.291592 23.290774 23.291634 23.292473 23.293294 23.294095 23.294878
#>  [99] 23.295643 23.296390
#> 
#> $fold
#>  [1]  2  3  9  1  9  6  3  8  8  4  4  1  1  5  7 10  2  7 10  6  5
#> 
#> $lambda
#>   [1] 5.508126876 5.136897560 4.790687857 4.467811529 4.166695985 3.885874621
#>   [7] 3.623979677 3.379735576 3.151952710 2.939521646 2.741407726 2.556646020
#>  [13] 2.384336635 2.223640326 2.073774410 1.934008956 1.803663225 1.682102360
#>  [19] 1.568734291 1.463006850 1.364405085 1.272448749 1.186689962 1.106711030
#>  [25] 1.032122411 0.962560815 0.897687437 0.837186309 0.780762753 0.728141957
#>  [31] 0.679067626 0.633300741 0.590618390 0.550812688 0.513689756 0.479068786
#>  [37] 0.446781153 0.416669599 0.388587462 0.362397968 0.337973558 0.315195271
#>  [43] 0.293952165 0.274140773 0.255664602 0.238433664 0.222364033 0.207377441
#>  [49] 0.193400896 0.180366323 0.168210236 0.156873429 0.146300685 0.136440509
#>  [55] 0.127244875 0.118668996 0.110671103 0.103212241 0.096256081 0.089768744
#>  [61] 0.083718631 0.078076275 0.072814196 0.067906763 0.063330074 0.059061839
#>  [67] 0.055081269 0.051368976 0.047906879 0.044678115 0.041666960 0.038858746
#>  [73] 0.036239797 0.033797356 0.031519527 0.029395216 0.027414077 0.025566460
#>  [79] 0.023843366 0.022236403 0.020737744 0.019340090 0.018036632 0.016821024
#>  [85] 0.015687343 0.014630069 0.013644051 0.012724487 0.011866900 0.011067110
#>  [91] 0.010321224 0.009625608 0.008976874 0.008371863 0.007807628 0.007281420
#>  [97] 0.006790676 0.006333007 0.005906184 0.005508127
#> 
#> $fit
#> $beta
#>               5.5081     5.1369     4.7907     4.4678    4.1667    3.8859
#> (Intercept) 20.32381 22.1589507 24.0115215 25.7392352 27.350507 28.853184
#> am           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> cyl          0.00000 -0.2964459 -0.5957073 -0.8747995 -1.135082 -1.377822
#> disp         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> wt           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#>                3.6240    3.3797    3.1520    2.9395    2.7414    2.5566
#> (Intercept) 30.254586 31.561538 32.780406 33.917126 34.977235 35.965896
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.604202 -1.815325 -2.012219 -2.195843 -2.367092 -2.526799
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                2.3843    2.2236   2.0738    1.9340    1.8037    1.6821
#> (Intercept) 36.887925 37.747813 38.54975 39.297632 39.872289 39.872289
#> am           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> cyl         -2.675742 -2.814647 -2.94419 -3.065002 -3.157831 -3.157831
#> disp         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#>                1.5687    1.4630    1.3644    1.2724    1.1867    1.1067
#> (Intercept) 39.872289 39.872289 39.872289 39.872289 39.872289 39.872289
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -3.157831 -3.157831 -3.157831 -3.157831 -3.157831 -3.157831
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.0321    0.9626    0.8977    0.8372    0.7808    0.7281
#> (Intercept) 39.872289 40.616822 40.616499 40.616281 40.616133 40.616034
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -3.157831 -1.692714 -1.692521 -1.692390 -1.692301 -1.692242
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000 -3.111043 -3.111321 -3.111509 -3.111636 -3.111721
#>                0.6791    0.6333    0.5906    0.5508    0.5137     0.4791
#> (Intercept) 40.615967 40.615921 40.615891 40.615870 40.615856 33.2052588
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> cyl         -1.692201 -1.692174 -1.692156 -1.692143 -1.692135 -1.2968845
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.3532982
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> wt          -3.111779 -3.111818 -3.111844 -3.111862 -3.111874 -3.5180423
#>                 0.4468    0.4167     0.3886     0.3624     0.3380     0.3152
#> (Intercept) -0.4621762 -0.465972 -0.4683719 -0.4698892 -0.4708486 -0.4714552
#> am           3.8419599  3.842749  3.8432483  3.8435638  3.8437633  3.8438895
#> carb         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         1.6909640  1.691099  1.6911839  1.6912378  1.6912718  1.6912934
#> vs           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -3.4660162 -3.465687 -3.4654791 -3.4653475 -3.4652644 -3.4652118
#>                 0.2940     0.2741     0.2557     0.2384     0.2224     0.2074
#> (Intercept) -0.4718387 -0.4720811 -0.4722344 -0.4723314 -0.4723927 -0.4724314
#> am           3.8439692  3.8440196  3.8440515  3.8440717  3.8440844  3.8440925
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         1.6913070  1.6913156  1.6913210  1.6913245  1.6913267  1.6913280
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -3.4651785 -3.4651575 -3.4651442 -3.4651358 -3.4651305 -3.4651272
#>                 0.1934     0.1804     0.1682     0.1569     0.1463     0.1364
#> (Intercept) -0.4724559 -0.4724714 -0.4724812 -0.4724874 -0.4724913 -0.4724938
#> am           3.8440976  3.8441008  3.8441028  3.8441041  3.8441049  3.8441054
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         1.6913289  1.6913295  1.6913298  1.6913300  1.6913302  1.6913302
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -3.4651250 -3.4651237 -3.4651228 -3.4651223 -3.4651220 -3.4651217
#>                 0.1272     0.1187     0.1107     0.1032     0.0963      0.0898
#> (Intercept) -0.4724953 -0.4724963 -0.4724969 -0.4724973 -0.4724976 -17.1059287
#> am           3.8441058  3.8441060  3.8441061  3.8441062  3.8441062   4.0093087
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  -0.3371691
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000   0.7889044
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000   0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000   0.5906766
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000   0.8759603
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000   0.0000000
#> qsec         1.6913303  1.6913303  1.6913304  1.6913304  1.6913304   2.1121100
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000   0.0000000
#> wt          -3.4651216 -3.4651215 -3.4651215 -3.4651214 -3.4651214  -3.5314192
#>                  0.0837      0.0781      0.0728      0.0679      0.0633
#> (Intercept) -17.1111948 -17.1157609 -17.1197333 -17.1231994 -17.1262320
#> am            4.0095197   4.0097316   4.0099367   4.0101306   4.0103107
#> carb         -0.3372699  -0.3373474  -0.3374069  -0.3374527  -0.3374880
#> cyl           0.7892043   0.7894637   0.7896886   0.7898843   0.7900550
#> disp          0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> drat          0.5906046   0.5905488   0.5905060   0.5904734   0.5904488
#> gear          0.8762559   0.8764838   0.8766604   0.8767980   0.8769058
#> hp            0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> qsec          2.1122865   2.1124427   2.1125809   2.1127033   2.1128117
#> vs            0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> wt           -3.5315307  -3.5316314  -3.5317224  -3.5318045  -3.5318787
#>                  0.0591      0.0551      0.0514      0.0479      0.0447
#> (Intercept) -17.1288919 -17.1312301 -17.1332895 -17.1351066 -17.1367125
#> am            4.0104760   4.0106264   4.0107624   4.0108849   4.0109949
#> carb         -0.3375153  -0.3375363  -0.3375527  -0.3375655  -0.3375754
#> cyl           0.7902044   0.7903355   0.7904508   0.7905525   0.7906423
#> disp          0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> drat          0.5904302   0.5904163   0.5904058   0.5903978   0.5903917
#> gear          0.8769910   0.8770587   0.8771130   0.8771570   0.8771928
#> hp            0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> qsec          2.1129078   2.1129931   2.1130688   2.1131361   2.1131959
#> vs            0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> wt           -3.5319457  -3.5320064  -3.5320614  -3.5321111  -3.5321561
#>                  0.0417      0.0389      0.0362      0.0338      0.0315
#> (Intercept) -17.1381337 -17.1393929 -17.1405100 -17.1415019 -17.1423834
#> am            4.0110935   4.0111817   4.0112607   4.0113313   4.0113945
#> carb         -0.3375832  -0.3375893  -0.3375941  -0.3375979  -0.3376009
#> cyl           0.7907219   0.7907924   0.7908550   0.7909106   0.7909601
#> disp          0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> drat          0.5903871   0.5903834   0.5903805   0.5903781   0.5903762
#> gear          0.8772221   0.8772464   0.8772666   0.8772835   0.8772977
#> hp            0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> qsec          2.1132492   2.1132966   2.1133388   2.1133765   2.1134101
#> vs            0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> wt           -3.5321969  -3.5322337  -3.5322671  -3.5322972  -3.5323244
#>                  0.0294      0.0274      0.0256      0.0238      0.0222
#> (Intercept) -17.1431675 -17.1438655 -17.1444870 -17.1450410 -17.1455348
#> am            4.0114510   4.0115015   4.0115467   4.0115872   4.0116235
#> carb         -0.3376033  -0.3376052  -0.3376067  -0.3376079  -0.3376089
#> cyl           0.7910042   0.7910435   0.7910785   0.7911097   0.7911376
#> disp          0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> drat          0.5903746   0.5903732   0.5903720   0.5903710   0.5903700
#> gear          0.8773097   0.8773199   0.8773286   0.8773360   0.8773423
#> hp            0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> qsec          2.1134401   2.1134669   2.1134908   2.1135122   2.1135312
#> vs            0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> wt           -3.5323490  -3.5323712  -3.5323911  -3.5324091  -3.5324253
#>                  0.0207      0.0193       0.0180        0.0168       0.0157
#> (Intercept) -17.1459754 -17.1463685 -17.87011983 -17.866451908 -17.86297683
#> am            4.0116559   4.0116850   3.99571225   3.995661494   3.99561188
#> carb         -0.3376096  -0.3376103  -0.27922503  -0.279002390  -0.27879477
#> cyl           0.7911625   0.7911848   0.74957195   0.749270788   0.74898762
#> disp          0.0000000   0.0000000   0.00243404   0.002438986   0.00244358
#> drat          0.5903692   0.5903685   0.59331307   0.593123943   0.59294727
#> gear          0.8773478   0.8773525   0.84773267   0.847380509   0.84705081
#> hp            0.0000000   0.0000000   0.00000000   0.000000000   0.00000000
#> qsec          2.1135483   2.1135636   2.17761169   2.177615463   2.17761681
#> vs            0.0000000   0.0000000   0.00000000   0.000000000   0.00000000
#> wt           -3.5324399  -3.5324530  -3.77175721  -3.772261401  -3.77272867
#>                    0.0146        0.0136       0.0127        0.0119
#> (Intercept) -17.859686989 -17.856574791 -17.85363273 -17.850853418
#> am            3.995563513   3.995516495   3.99547089   3.995426761
#> carb         -0.278601237  -0.278420904  -0.27825294  -0.278096544
#> cyl           0.748721513   0.748471565   0.74823691   0.748016733
#> disp          0.002447846   0.002451806   0.00245548   0.002458889
#> drat          0.592782313   0.592628347   0.59248470   0.592350732
#> gear          0.846742265   0.846453650   0.84618379   0.845931570
#> hp            0.000000000   0.000000000   0.00000000   0.000000000
#> qsec          2.177616059   2.177613500   2.17760940   2.177603983
#> vs            0.000000000   0.000000000   0.00000000   0.000000000
#> wt           -3.773161570  -3.773562465  -3.77393358  -3.774277003
#>                    0.0111        0.0103        0.0096        0.0090
#> (Intercept) -17.848229575 -17.845754092 -17.843420023 -17.841220611
#> am            3.995384137   3.995343046   3.995303500   3.995265502
#> carb         -0.277950987  -0.277815566  -0.277689622  -0.277572536
#> cyl           0.747810226   0.747616637   0.747435240   0.747265345
#> disp          0.002462049   0.002464978   0.002467692   0.002470206
#> drat          0.592225841   0.592109457   0.592001042   0.591900090
#> gear          0.845695938   0.845475892   0.845270485   0.845078820
#> hp            0.000000000   0.000000000   0.000000000   0.000000000
#> qsec          2.177597471   2.177590047   2.177581877   2.177573109
#> vs            0.000000000   0.000000000   0.000000000   0.000000000
#> wt           -3.774594669  -3.774888398  -3.775159887  -3.775410719
#>                    0.0084        0.0078        0.0073        0.0068
#> (Intercept) -17.839149294 -17.515476119 -17.515697299 -17.515814516
#> am            3.995229047   3.976371847   3.976426980   3.976479847
#> carb         -0.277463726  -0.263535286  -0.263359027  -0.263190557
#> cyl           0.747106293   0.740617954   0.740509411   0.740400843
#> disp          0.002472533   0.002847661   0.002852583   0.002857252
#> drat          0.591806122   0.581540392   0.581404384   0.581274010
#> gear          0.844900047   0.865696929   0.865513077   0.865329982
#> hp            0.000000000  -0.001081616  -0.001081655  -0.001081612
#> qsec          2.177563873   2.161951891   2.162077982   2.162195746
#> vs            0.000000000   0.000000000   0.000000000   0.000000000
#> wt           -3.775642374  -3.776223131  -3.776787668  -3.777324387
#>                    0.0063        0.0059        0.0055
#> (Intercept) -17.515836520 -17.515771884 -17.515628932
#> am            3.976530225   3.976577955   3.976622933
#> carb         -0.263029709  -0.262876302  -0.262730146
#> cyl           0.740292662   0.740185244   0.740078929
#> disp          0.002861677   0.002865865   0.002869826
#> drat          0.581149173   0.581029763   0.580915663
#> gear          0.865148519   0.864969449   0.864793425
#> hp           -0.001081501  -0.001081334  -0.001081123
#> qsec          2.162305472   2.162407470   2.162502066
#> vs            0.000000000   0.000000000   0.000000000
#> wt           -3.777833905  -3.778316917  -3.778774179
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1   1   1  50   1   1   1   1   1   1   1   1   1 103
#>  [37] 125   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [55]   1   1   1   1   1 136   1   1   1   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1  61   1   1   1   1   1   1   1
#>  [91]   1   1   1   1  47   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.508126876 5.136897560 4.790687857 4.467811529 4.166695985 3.885874621
#>   [7] 3.623979677 3.379735576 3.151952710 2.939521646 2.741407726 2.556646020
#>  [13] 2.384336635 2.223640326 2.073774410 1.934008956 1.803663225 1.682102360
#>  [19] 1.568734291 1.463006850 1.364405085 1.272448749 1.186689962 1.106711030
#>  [25] 1.032122411 0.962560815 0.897687437 0.837186309 0.780762753 0.728141957
#>  [31] 0.679067626 0.633300741 0.590618390 0.550812688 0.513689756 0.479068786
#>  [37] 0.446781153 0.416669599 0.388587462 0.362397968 0.337973558 0.315195271
#>  [43] 0.293952165 0.274140773 0.255664602 0.238433664 0.222364033 0.207377441
#>  [49] 0.193400896 0.180366323 0.168210236 0.156873429 0.146300685 0.136440509
#>  [55] 0.127244875 0.118668996 0.110671103 0.103212241 0.096256081 0.089768744
#>  [61] 0.083718631 0.078076275 0.072814196 0.067906763 0.063330074 0.059061839
#>  [67] 0.055081269 0.051368976 0.047906879 0.044678115 0.041666960 0.038858746
#>  [73] 0.036239797 0.033797356 0.031519527 0.029395216 0.027414077 0.025566460
#>  [79] 0.023843366 0.022236403 0.020737744 0.019340090 0.018036632 0.016821024
#>  [85] 0.015687343 0.014630069 0.013644051 0.012724487 0.011866900 0.011067110
#>  [91] 0.010321224 0.009625608 0.008976874 0.008371863 0.007807628 0.007281420
#>  [97] 0.006790676 0.006333007 0.005906184 0.005508127
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 25
#> 
#> $loss
#>   [1] 853.8981 741.0578 638.4194 552.9060 482.0339 423.6599 375.9360 337.2696
#>   [9] 306.2896 281.8161 262.8350 248.4758 237.9914 230.7421 226.1806 223.8392
#>  [17] 223.2942 223.2942 223.2942 223.2942 223.2942 223.2942 223.2942 223.2942
#>  [25] 223.2942 157.9849 157.9849 157.9849 157.9849 157.9849 157.9849 157.9849
#>  [33] 157.9849 157.9849 157.9849 148.7907 116.1113 116.1113 116.1113 116.1113
#>  [41] 116.1113 116.1113 116.1113 116.1113 116.1113 116.1113 116.1113 116.1113
#>  [49] 116.1113 116.1113 116.1113 116.1113 116.1113 116.1113 116.1113 116.1113
#>  [57] 116.1113 116.1113 116.1113 111.6096 111.6096 111.6096 111.6096 111.6096
#>  [65] 111.6096 111.6096 111.6096 111.6096 111.6096 111.6096 111.6096 111.6096
#>  [73] 111.6096 111.6096 111.6096 111.6096 111.6096 111.6096 111.6096 111.6096
#>  [81] 111.6096 111.6096 111.4930 111.4929 111.4929 111.4929 111.4929 111.4929
#>  [89] 111.4929 111.4928 111.4928 111.4928 111.4928 111.4928 111.4793 111.4793
#>  [97] 111.4793 111.4793 111.4793 111.4792
#> 
#> $linear.predictors
#>      5.5081   5.1369   4.7907   4.4678   4.1667   3.8859   3.6240   3.3797
#> 1  20.32381 20.38028 20.43728 20.49044 20.54002 20.58625 20.62937 20.66959
#> 2  20.32381 20.38028 20.43728 20.49044 20.54002 20.58625 20.62937 20.66959
#> 3  20.32381 20.97317 21.62869 22.24004 22.81018 23.34190 23.83778 24.30024
#> 4  20.32381 20.38028 20.43728 20.49044 20.54002 20.58625 20.62937 20.66959
#> 5  20.32381 20.38028 20.43728 20.49044 20.54002 20.58625 20.62937 20.66959
#> 6  20.32381 20.38028 20.43728 20.49044 20.54002 20.58625 20.62937 20.66959
#> 7  20.32381 19.78738 19.24586 18.74084 18.26985 17.83061 17.42097 17.03894
#> 8  20.32381 19.78738 19.24586 18.74084 18.26985 17.83061 17.42097 17.03894
#> 9  20.32381 19.78738 19.24586 18.74084 18.26985 17.83061 17.42097 17.03894
#> 10 20.32381 19.78738 19.24586 18.74084 18.26985 17.83061 17.42097 17.03894
#> 11 20.32381 20.97317 21.62869 22.24004 22.81018 23.34190 23.83778 24.30024
#> 12 20.32381 20.97317 21.62869 22.24004 22.81018 23.34190 23.83778 24.30024
#> 13 20.32381 20.97317 21.62869 22.24004 22.81018 23.34190 23.83778 24.30024
#> 14 20.32381 20.97317 21.62869 22.24004 22.81018 23.34190 23.83778 24.30024
#> 15 20.32381 19.78738 19.24586 18.74084 18.26985 17.83061 17.42097 17.03894
#> 16 20.32381 19.78738 19.24586 18.74084 18.26985 17.83061 17.42097 17.03894
#> 17 20.32381 19.78738 19.24586 18.74084 18.26985 17.83061 17.42097 17.03894
#> 18 20.32381 20.97317 21.62869 22.24004 22.81018 23.34190 23.83778 24.30024
#> 19 20.32381 19.78738 19.24586 18.74084 18.26985 17.83061 17.42097 17.03894
#> 20 20.32381 19.78738 19.24586 18.74084 18.26985 17.83061 17.42097 17.03894
#> 21 20.32381 20.97317 21.62869 22.24004 22.81018 23.34190 23.83778 24.30024
#>      3.1520   2.9395   2.7414   2.5566   2.3843   2.2236   2.0738   1.9340
#> 1  20.70709 20.74207 20.77468 20.80510 20.83347 20.85993 20.88461 20.90762
#> 2  20.70709 20.74207 20.77468 20.80510 20.83347 20.85993 20.88461 20.90762
#> 3  24.73153 25.13375 25.50887 25.85870 26.18496 26.48923 26.77299 27.03762
#> 4  20.70709 20.74207 20.77468 20.80510 20.83347 20.85993 20.88461 20.90762
#> 5  20.70709 20.74207 20.77468 20.80510 20.83347 20.85993 20.88461 20.90762
#> 6  20.70709 20.74207 20.77468 20.80510 20.83347 20.85993 20.88461 20.90762
#> 7  16.68265 16.35038 16.04050 15.75151 15.48199 15.23064 14.99623 14.77762
#> 8  16.68265 16.35038 16.04050 15.75151 15.48199 15.23064 14.99623 14.77762
#> 9  16.68265 16.35038 16.04050 15.75151 15.48199 15.23064 14.99623 14.77762
#> 10 16.68265 16.35038 16.04050 15.75151 15.48199 15.23064 14.99623 14.77762
#> 11 24.73153 25.13375 25.50887 25.85870 26.18496 26.48923 26.77299 27.03762
#> 12 24.73153 25.13375 25.50887 25.85870 26.18496 26.48923 26.77299 27.03762
#> 13 24.73153 25.13375 25.50887 25.85870 26.18496 26.48923 26.77299 27.03762
#> 14 24.73153 25.13375 25.50887 25.85870 26.18496 26.48923 26.77299 27.03762
#> 15 16.68265 16.35038 16.04050 15.75151 15.48199 15.23064 14.99623 14.77762
#> 16 16.68265 16.35038 16.04050 15.75151 15.48199 15.23064 14.99623 14.77762
#> 17 16.68265 16.35038 16.04050 15.75151 15.48199 15.23064 14.99623 14.77762
#> 18 24.73153 25.13375 25.50887 25.85870 26.18496 26.48923 26.77299 27.03762
#> 19 16.68265 16.35038 16.04050 15.75151 15.48199 15.23064 14.99623 14.77762
#> 20 16.68265 16.35038 16.04050 15.75151 15.48199 15.23064 14.99623 14.77762
#> 21 24.73153 25.13375 25.50887 25.85870 26.18496 26.48923 26.77299 27.03762
#>      1.8037   1.6821   1.5687   1.4630   1.3644   1.2724   1.1867   1.1067
#> 1  20.92530 20.92530 20.92530 20.92530 20.92530 20.92530 20.92530 20.92530
#> 2  20.92530 20.92530 20.92530 20.92530 20.92530 20.92530 20.92530 20.92530
#> 3  27.24096 27.24096 27.24096 27.24096 27.24096 27.24096 27.24096 27.24096
#> 4  20.92530 20.92530 20.92530 20.92530 20.92530 20.92530 20.92530 20.92530
#> 5  20.92530 20.92530 20.92530 20.92530 20.92530 20.92530 20.92530 20.92530
#> 6  20.92530 20.92530 20.92530 20.92530 20.92530 20.92530 20.92530 20.92530
#> 7  14.60964 14.60964 14.60964 14.60964 14.60964 14.60964 14.60964 14.60964
#> 8  14.60964 14.60964 14.60964 14.60964 14.60964 14.60964 14.60964 14.60964
#> 9  14.60964 14.60964 14.60964 14.60964 14.60964 14.60964 14.60964 14.60964
#> 10 14.60964 14.60964 14.60964 14.60964 14.60964 14.60964 14.60964 14.60964
#> 11 27.24096 27.24096 27.24096 27.24096 27.24096 27.24096 27.24096 27.24096
#> 12 27.24096 27.24096 27.24096 27.24096 27.24096 27.24096 27.24096 27.24096
#> 13 27.24096 27.24096 27.24096 27.24096 27.24096 27.24096 27.24096 27.24096
#> 14 27.24096 27.24096 27.24096 27.24096 27.24096 27.24096 27.24096 27.24096
#> 15 14.60964 14.60964 14.60964 14.60964 14.60964 14.60964 14.60964 14.60964
#> 16 14.60964 14.60964 14.60964 14.60964 14.60964 14.60964 14.60964 14.60964
#> 17 14.60964 14.60964 14.60964 14.60964 14.60964 14.60964 14.60964 14.60964
#> 18 27.24096 27.24096 27.24096 27.24096 27.24096 27.24096 27.24096 27.24096
#> 19 14.60964 14.60964 14.60964 14.60964 14.60964 14.60964 14.60964 14.60964
#> 20 14.60964 14.60964 14.60964 14.60964 14.60964 14.60964 14.60964 14.60964
#> 21 27.24096 27.24096 27.24096 27.24096 27.24096 27.24096 27.24096 27.24096
#>      1.0321   0.9626   0.8977   0.8372   0.7808   0.7281   0.6791   0.6333
#> 1  20.92530 22.30960 22.30971 22.30979 22.30984 22.30987 22.30990 22.30991
#> 2  20.92530 21.51629 21.51633 21.51635 21.51637 21.51639 21.51639 21.51640
#> 3  27.24096 26.62834 26.62815 26.62802 26.62793 26.62787 26.62783 26.62781
#> 4  20.92530 20.45853 20.45848 20.45844 20.45842 20.45840 20.45839 20.45838
#> 5  20.92530 19.75855 19.75843 19.75835 19.75830 19.75826 19.75824 19.75822
#> 6  20.92530 19.75855 19.75843 19.75835 19.75830 19.75826 19.75824 19.75822
#> 7  14.60964 14.41316 14.41326 14.41332 14.41337 14.41340 14.41342 14.41343
#> 8  14.60964 15.47092 15.47111 15.47124 15.47132 15.47138 15.47142 15.47145
#> 9  14.60964 10.74213 10.74190 10.74174 10.74164 10.74157 10.74152 10.74148
#> 10 14.60964 10.44658 10.44632 10.44615 10.44603 10.44595 10.44590 10.44586
#> 11 27.24096 27.00167 27.00151 27.00140 27.00133 27.00128 27.00125 27.00123
#> 12 27.24096 28.82163 28.82163 28.82164 28.82164 28.82164 28.82164 28.82164
#> 13 27.24096 28.13720 28.13714 28.13710 28.13708 28.13706 28.13705 28.13704
#> 14 27.24096 26.17724 26.17701 26.17685 26.17675 26.17667 26.17663 26.17659
#> 15 14.60964 16.12423 16.12448 16.12465 16.12477 16.12484 16.12489 16.12493
#> 16 14.60964 16.38867 16.38895 16.38913 16.38925 16.38934 16.38940 16.38943
#> 17 14.60964 15.12870 15.12886 15.12897 15.12904 15.12909 15.12913 15.12915
#> 18 27.24096 29.13896 29.13899 29.13901 29.13902 29.13903 29.13904 29.13904
#> 19 14.60964 17.21310 17.21345 17.21368 17.21384 17.21394 17.21402 17.21407
#> 20 14.60964 15.96868 15.96892 15.96908 15.96918 15.96926 15.96931 15.96934
#> 21 27.24096 25.19726 25.19694 25.19673 25.19658 25.19648 25.19642 25.19637
#>      0.5906   0.5508   0.5137   0.4791   0.4468   0.4167   0.3886   0.3624
#> 1  22.30992 22.30993 22.30994 22.02197 22.13209 22.13216 22.13221 22.13224
#> 2  21.51640 21.51641 21.51641 21.32272 22.19519 22.19543 22.19557 22.19567
#> 3  26.62779 26.62778 26.62777 26.43074 26.80747 26.80773 26.80790 26.80800
#> 4  20.45838 20.45837 20.45837 20.98156 21.26692 21.26680 21.26673 21.26668
#> 5  19.75821 19.75820 19.75820 19.78724 18.55937 18.55917 18.55905 18.55897
#> 6  19.75821 19.75820 19.75820 19.99922 19.57395 19.57383 19.57376 19.57371
#> 7  14.41344 14.41344 14.41345 14.65914 14.85391 14.85380 14.85373 14.85368
#> 8  15.47147 15.47148 15.47149 15.92593 16.37055 16.37035 16.37023 16.37015
#> 9  10.74146 10.74145 10.74144 10.71276 11.74477 11.74513 11.74535 11.74549
#> 10 10.44584 10.44582 10.44581 10.18070 10.46856 10.46887 10.46907 10.46919
#> 11 27.00121 27.00120 27.00119 27.15674 28.67762 28.67796 28.67817 28.67831
#> 12 28.82164 28.82164 28.82164 28.87916 29.09882 29.09884 29.09885 29.09886
#> 13 28.13703 28.13703 28.13703 28.59275 30.66983 30.67011 30.67028 30.67039
#> 14 26.17657 26.17656 26.17655 26.41524 24.83028 24.82999 24.82981 24.82970
#> 15 16.12495 16.12497 16.12498 16.40681 15.86401 15.86364 15.86341 15.86327
#> 16 16.38946 16.38948 16.38949 16.85777 16.88574 16.88540 16.88519 16.88506
#> 17 15.12916 15.12917 15.12918 14.76523 12.28608 12.28562 12.28533 12.28515
#> 18 29.13905 29.13905 29.13905 28.66566 26.71299 26.71276 26.71261 26.71252
#> 19 17.21410 17.21412 17.21414 16.80081 16.91149 16.91148 16.91147 16.91147
#> 20 15.96936 15.96938 15.96939 15.42893 15.69418 15.69432 15.69440 15.69446
#> 21 25.19634 25.19632 25.19631 24.80891 25.19619 25.19660 25.19687 25.19703
#>      0.3380   0.3152   0.2940   0.2741   0.2557   0.2384   0.2224   0.2074
#> 1  22.13226 22.13227 22.13228 22.13228 22.13228 22.13229 22.13229 22.13229
#> 2  22.19573 22.19576 22.19579 22.19580 22.19581 22.19582 22.19582 22.19582
#> 3  26.80807 26.80811 26.80814 26.80816 26.80817 26.80817 26.80818 26.80818
#> 4  21.26665 21.26663 21.26662 21.26661 21.26661 21.26660 21.26660 21.26660
#> 5  18.55892 18.55889 18.55887 18.55885 18.55884 18.55884 18.55884 18.55883
#> 6  19.57368 19.57366 19.57365 19.57364 19.57364 19.57363 19.57363 19.57363
#> 7  14.85366 14.85364 14.85363 14.85362 14.85361 14.85361 14.85361 14.85361
#> 8  16.37010 16.37007 16.37005 16.37004 16.37003 16.37002 16.37002 16.37002
#> 9  11.74558 11.74564 11.74567 11.74570 11.74571 11.74572 11.74573 11.74573
#> 10 10.46927 10.46932 10.46935 10.46937 10.46938 10.46939 10.46940 10.46940
#> 11 28.67840 28.67845 28.67848 28.67851 28.67852 28.67853 28.67853 28.67854
#> 12 29.09887 29.09887 29.09887 29.09887 29.09887 29.09888 29.09888 29.09888
#> 13 30.67046 30.67051 30.67054 30.67055 30.67057 30.67057 30.67058 30.67058
#> 14 24.82962 24.82958 24.82955 24.82953 24.82952 24.82951 24.82951 24.82950
#> 15 15.86318 15.86312 15.86308 15.86306 15.86304 15.86303 15.86303 15.86302
#> 16 16.88497 16.88492 16.88488 16.88486 16.88485 16.88484 16.88484 16.88483
#> 17 12.28504 12.28496 12.28492 12.28489 12.28487 12.28486 12.28485 12.28485
#> 18 26.71246 26.71243 26.71240 26.71239 26.71238 26.71237 26.71237 26.71237
#> 19 16.91147 16.91147 16.91147 16.91147 16.91146 16.91146 16.91146 16.91146
#> 20 15.69449 15.69451 15.69453 15.69453 15.69454 15.69454 15.69455 15.69455
#> 21 25.19714 25.19720 25.19724 25.19727 25.19729 25.19730 25.19730 25.19731
#>      0.1934   0.1804   0.1682   0.1569   0.1463   0.1364   0.1272   0.1187
#> 1  22.13229 22.13229 22.13229 22.13229 22.13229 22.13229 22.13229 22.13229
#> 2  22.19583 22.19583 22.19583 22.19583 22.19583 22.19583 22.19583 22.19583
#> 3  26.80818 26.80818 26.80818 26.80818 26.80818 26.80819 26.80819 26.80819
#> 4  21.26660 21.26660 21.26660 21.26660 21.26660 21.26660 21.26660 21.26660
#> 5  18.55883 18.55883 18.55883 18.55883 18.55883 18.55883 18.55883 18.55883
#> 6  19.57363 19.57363 19.57363 19.57363 19.57363 19.57363 19.57363 19.57363
#> 7  14.85361 14.85361 14.85361 14.85361 14.85361 14.85361 14.85361 14.85361
#> 8  16.37002 16.37002 16.37002 16.37001 16.37001 16.37001 16.37001 16.37001
#> 9  11.74573 11.74573 11.74573 11.74573 11.74573 11.74573 11.74574 11.74574
#> 10 10.46940 10.46940 10.46940 10.46940 10.46940 10.46940 10.46940 10.46940
#> 11 28.67854 28.67854 28.67854 28.67854 28.67854 28.67854 28.67854 28.67854
#> 12 29.09888 29.09888 29.09888 29.09888 29.09888 29.09888 29.09888 29.09888
#> 13 30.67058 30.67058 30.67058 30.67058 30.67058 30.67059 30.67059 30.67059
#> 14 24.82950 24.82950 24.82950 24.82950 24.82950 24.82950 24.82950 24.82950
#> 15 15.86302 15.86302 15.86302 15.86302 15.86302 15.86302 15.86302 15.86302
#> 16 16.88483 16.88483 16.88483 16.88483 16.88483 16.88483 16.88483 16.88483
#> 17 12.28484 12.28484 12.28484 12.28484 12.28484 12.28484 12.28484 12.28484
#> 18 26.71237 26.71236 26.71236 26.71236 26.71236 26.71236 26.71236 26.71236
#> 19 16.91146 16.91146 16.91146 16.91146 16.91146 16.91146 16.91146 16.91146
#> 20 15.69455 15.69455 15.69455 15.69455 15.69455 15.69455 15.69455 15.69455
#> 21 25.19731 25.19731 25.19731 25.19732 25.19732 25.19732 25.19732 25.19732
#>      0.1107   0.1032   0.0963   0.0898   0.0837   0.0781   0.0728   0.0679
#> 1  22.13229 22.13229 22.13229 21.60862 21.60848 21.60837 21.60829 21.60823
#> 2  22.19583 22.19583 22.19583 21.89089 21.89082 21.89077 21.89075 21.89074
#> 3  26.80819 26.80819 26.80819 26.61325 26.61322 26.61320 26.61318 26.61315
#> 4  21.26660 21.26660 21.26660 21.44340 21.44357 21.44371 21.44382 21.44391
#> 5  18.55883 18.55883 18.55883 18.60165 18.60152 18.60141 18.60130 18.60121
#> 6  19.57363 19.57363 19.57363 19.86891 19.86889 19.86887 19.86885 19.86883
#> 7  14.85361 14.85361 14.85361 15.01290 15.01301 15.01311 15.01319 15.01326
#> 8  16.37001 16.37001 16.37001 16.63600 16.63619 16.63635 16.63649 16.63661
#> 9  11.74574 11.74574 11.74574 11.65098 11.65097 11.65097 11.65097 11.65098
#> 10 10.46940 10.46940 10.46940 10.30992 10.30978 10.30966 10.30957 10.30948
#> 11 28.67854 28.67854 28.67854 28.98929 28.98941 28.98952 28.98962 28.98970
#> 12 29.09888 29.09888 29.09888 29.21357 29.21343 29.21333 29.21325 29.21319
#> 13 30.67059 30.67059 30.67059 31.26916 31.26939 31.26960 31.26978 31.26994
#> 14 24.82950 24.82950 24.82950 24.08428 24.08399 24.08374 24.08352 24.08333
#> 15 15.86302 15.86302 15.86302 15.98982 15.99002 15.99019 15.99032 15.99042
#> 16 16.88483 16.88483 16.88483 17.42856 17.42882 17.42904 17.42922 17.42937
#> 17 12.28484 12.28484 12.28484 11.67470 11.67434 11.67403 11.67377 11.67355
#> 18 26.71236 26.71236 26.71236 26.34293 26.34290 26.34284 26.34278 26.34270
#> 19 16.91146 16.91146 16.91146 18.16939 18.16971 18.16997 18.17019 18.17036
#> 20 15.69455 15.69455 15.69455 15.21770 15.21764 15.21760 15.21758 15.21758
#> 21 25.19732 25.19732 25.19732 24.78408 24.78388 24.78372 24.78358 24.78347
#>      0.0633   0.0591   0.0551   0.0514   0.0479   0.0447   0.0417   0.0389
#> 1  21.60819 21.60816 21.60813 21.60811 21.60810 21.60809 21.60808 21.60808
#> 2  21.89073 21.89074 21.89075 21.89076 21.89077 21.89078 21.89079 21.89080
#> 3  26.61313 26.61311 26.61309 26.61307 26.61305 26.61303 26.61301 26.61300
#> 4  21.44398 21.44404 21.44409 21.44413 21.44417 21.44420 21.44422 21.44425
#> 5  18.60112 18.60104 18.60098 18.60092 18.60086 18.60081 18.60077 18.60074
#> 6  19.86881 19.86879 19.86877 19.86876 19.86874 19.86873 19.86872 19.86871
#> 7  15.01332 15.01337 15.01341 15.01345 15.01348 15.01351 15.01354 15.01356
#> 8  16.63672 16.63681 16.63689 16.63697 16.63703 16.63709 16.63714 16.63718
#> 9  11.65098 11.65098 11.65099 11.65099 11.65099 11.65099 11.65100 11.65100
#> 10 10.30941 10.30935 10.30929 10.30925 10.30920 10.30917 10.30913 10.30910
#> 11 28.98978 28.98984 28.98990 28.98995 28.98999 28.99003 28.99006 28.99009
#> 12 29.21315 29.21312 29.21309 29.21308 29.21307 29.21306 29.21305 29.21305
#> 13 31.27008 31.27021 31.27032 31.27042 31.27051 31.27059 31.27066 31.27073
#> 14 24.08316 24.08301 24.08289 24.08278 24.08268 24.08259 24.08252 24.08245
#> 15 15.99051 15.99058 15.99064 15.99069 15.99073 15.99076 15.99079 15.99082
#> 16 17.42950 17.42961 17.42971 17.42979 17.42986 17.42992 17.42997 17.43002
#> 17 11.67336 11.67320 11.67306 11.67293 11.67283 11.67273 11.67265 11.67258
#> 18 26.34263 26.34256 26.34249 26.34242 26.34236 26.34230 26.34225 26.34220
#> 19 18.17050 18.17062 18.17072 18.17081 18.17088 18.17094 18.17099 18.17104
#> 20 15.21758 15.21758 15.21759 15.21760 15.21762 15.21763 15.21765 15.21766
#> 21 24.78337 24.78328 24.78320 24.78314 24.78308 24.78303 24.78299 24.78295
#>      0.0362   0.0338   0.0315   0.0294   0.0274   0.0256   0.0238   0.0222
#> 1  21.60807 21.60807 21.60807 21.60806 21.60806 21.60806 21.60806 21.60806
#> 2  21.89081 21.89082 21.89083 21.89084 21.89085 21.89086 21.89086 21.89087
#> 3  26.61298 26.61297 26.61296 26.61295 26.61294 26.61293 26.61292 26.61291
#> 4  21.44427 21.44428 21.44430 21.44431 21.44432 21.44433 21.44434 21.44435
#> 5  18.60070 18.60067 18.60065 18.60062 18.60060 18.60059 18.60057 18.60055
#> 6  19.86871 19.86870 19.86869 19.86869 19.86868 19.86868 19.86868 19.86867
#> 7  15.01358 15.01360 15.01362 15.01363 15.01364 15.01366 15.01367 15.01368
#> 8  16.63722 16.63726 16.63729 16.63732 16.63734 16.63737 16.63739 16.63741
#> 9  11.65100 11.65100 11.65100 11.65100 11.65100 11.65100 11.65100 11.65100
#> 10 10.30908 10.30905 10.30903 10.30901 10.30900 10.30898 10.30897 10.30895
#> 11 28.99011 28.99014 28.99016 28.99017 28.99019 28.99020 28.99022 28.99023
#> 12 29.21305 29.21305 29.21305 29.21305 29.21305 29.21305 29.21305 29.21305
#> 13 31.27078 31.27083 31.27087 31.27091 31.27095 31.27098 31.27101 31.27103
#> 14 24.08239 24.08234 24.08230 24.08226 24.08222 24.08219 24.08216 24.08213
#> 15 15.99084 15.99086 15.99087 15.99089 15.99090 15.99091 15.99092 15.99093
#> 16 17.43006 17.43010 17.43013 17.43016 17.43019 17.43021 17.43023 17.43025
#> 17 11.67252 11.67246 11.67241 11.67237 11.67233 11.67230 11.67227 11.67224
#> 18 26.34216 26.34212 26.34208 26.34205 26.34202 26.34199 26.34197 26.34195
#> 19 18.17108 18.17111 18.17114 18.17117 18.17119 18.17121 18.17123 18.17124
#> 20 15.21767 15.21768 15.21769 15.21770 15.21771 15.21772 15.21773 15.21774
#> 21 24.78291 24.78288 24.78285 24.78283 24.78280 24.78278 24.78277 24.78275
#>      0.0207   0.0193   0.0180   0.0168   0.0157   0.0146   0.0136   0.0127
#> 1  21.60806 21.60806 21.56191 21.56199 21.56208 21.56215 21.56223 21.56229
#> 2  21.89088 21.89088 21.81957 21.81953 21.81950 21.81946 21.81943 21.81940
#> 3  26.61291 26.61290 26.55759 26.55753 26.55747 26.55741 26.55736 26.55731
#> 4  21.44436 21.44436 21.55324 21.55342 21.55358 21.55372 21.55386 21.55398
#> 5  18.60054 18.60053 18.51052 18.51029 18.51007 18.50987 18.50968 18.50951
#> 6  19.86867 19.86867 19.81709 19.81686 19.81664 19.81644 19.81625 19.81608
#> 7  15.01368 15.01369 14.86415 14.86382 14.86351 14.86322 14.86295 14.86270
#> 8  16.63742 16.63744 16.58207 16.58191 16.58176 16.58162 16.58148 16.58136
#> 9  11.65100 11.65100 11.79176 11.79205 11.79233 11.79258 11.79282 11.79303
#> 10 10.30894 10.30893 10.31409 10.31411 10.31414 10.31417 10.31419 10.31422
#> 11 28.99024 28.99025 28.94810 28.94790 28.94772 28.94756 28.94740 28.94725
#> 12 29.21305 29.21305 29.30363 29.30378 29.30391 29.30404 29.30416 29.30427
#> 13 31.27105 31.27107 31.32573 31.32566 31.32559 31.32552 31.32546 31.32540
#> 14 24.08211 24.08209 24.15636 24.15671 24.15705 24.15736 24.15765 24.15793
#> 15 15.99094 15.99095 15.98250 15.98248 15.98247 15.98246 15.98245 15.98243
#> 16 17.43027 17.43028 17.43678 17.43667 17.43657 17.43647 17.43637 17.43628
#> 17 11.67222 11.67220 11.69117 11.69141 11.69164 11.69185 11.69205 11.69224
#> 18 26.34193 26.34191 26.36733 26.36749 26.36763 26.36777 26.36790 26.36802
#> 19 18.17126 18.17127 18.22096 18.22069 18.22044 18.22020 18.21998 18.21977
#> 20 15.21774 15.21775 15.28796 15.28827 15.28855 15.28881 15.28905 15.28928
#> 21 24.78274 24.78272 24.70749 24.70743 24.70737 24.70732 24.70728 24.70724
#>      0.0119   0.0111   0.0103   0.0096   0.0090   0.0084   0.0078   0.0073
#> 1  21.56236 21.56242 21.56248 21.56253 21.56258 21.56263 21.60993 21.60993
#> 2  21.81938 21.81935 21.81933 21.81931 21.81929 21.81928 21.85769 21.85761
#> 3  26.55727 26.55723 26.55719 26.55716 26.55713 26.55710 26.54159 26.54148
#> 4  21.55410 21.55421 21.55430 21.55439 21.55448 21.55455 21.55644 21.55667
#> 5  18.50935 18.50920 18.50907 18.50894 18.50883 18.50872 18.53426 18.53401
#> 6  19.81591 19.81576 19.81562 19.81549 19.81537 19.81526 19.83143 19.83125
#> 7  14.86247 14.86225 14.86205 14.86186 14.86169 14.86153 14.84071 14.84043
#> 8  16.58124 16.58113 16.58103 16.58093 16.58084 16.58076 16.55702 16.55695
#> 9  11.79324 11.79343 11.79360 11.79376 11.79391 11.79405 11.82542 11.82570
#> 10 10.31424 10.31427 10.31429 10.31431 10.31433 10.31435 10.31228 10.31224
#> 11 28.94712 28.94699 28.94687 28.94676 28.94665 28.94655 28.93354 28.93343
#> 12 29.30437 29.30446 29.30455 29.30463 29.30470 29.30477 29.32615 29.32629
#> 13 31.32534 31.32528 31.32523 31.32518 31.32513 31.32508 31.30236 31.30244
#> 14 24.15818 24.15843 24.15865 24.15886 24.15906 24.15924 24.12161 24.12179
#> 15 15.98242 15.98241 15.98240 15.98239 15.98238 15.98237 16.00768 16.00771
#> 16 17.43620 17.43612 17.43604 17.43597 17.43590 17.43584 17.44523 17.44524
#> 17 11.69242 11.69259 11.69275 11.69289 11.69303 11.69316 11.66823 11.66828
#> 18 26.36813 26.36824 26.36834 26.36844 26.36852 26.36861 26.38934 26.38940
#> 19 18.21957 18.21938 18.21921 18.21904 18.21889 18.21875 18.20594 18.20587
#> 20 15.28949 15.28968 15.28987 15.29003 15.29019 15.29034 15.24288 15.24315
#> 21 24.70720 24.70717 24.70715 24.70712 24.70710 24.70708 24.69029 24.69012
#>      0.0068   0.0063   0.0059   0.0055
#> 1  21.60993 21.60993 21.60993 21.60994
#> 2  21.85754 21.85748 21.85741 21.85735
#> 3  26.54137 26.54126 26.54117 26.54108
#> 4  21.55690 21.55710 21.55730 21.55749
#> 5  18.53377 18.53353 18.53331 18.53310
#> 6  19.83108 19.83092 19.83075 19.83060
#> 7  14.84016 14.83990 14.83966 14.83942
#> 8  16.55689 16.55683 16.55677 16.55671
#> 9  11.82597 11.82623 11.82647 11.82670
#> 10 10.31221 10.31218 10.31215 10.31212
#> 11 28.93331 28.93321 28.93310 28.93300
#> 12 29.32642 29.32655 29.32667 29.32678
#> 13 31.30253 31.30260 31.30267 31.30273
#> 14 24.12197 24.12216 24.12233 24.12251
#> 15 16.00774 16.00777 16.00779 16.00781
#> 16 17.44525 17.44526 17.44526 17.44526
#> 17 11.66833 11.66838 11.66844 11.66851
#> 18 26.38946 26.38952 26.38957 26.38962
#> 19 18.20581 18.20574 18.20567 18.20559
#> 20 15.24341 15.24366 15.24391 15.24414
#> 21 24.68995 24.68980 24.68967 24.68954
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 22.8 21.4 19.2 17.8 16.4 17.3 10.4 14.7 32.4 30.4 33.9 21.5 15.5
#> [16] 15.2 13.3 30.4 15.8 15.0 21.4
#> 
#> $X
#>               am        carb        cyl       disp        drat       gear
#>  [1,]  1.0488088  0.66011331 -0.1097643 -0.5218446  0.43838004  0.4082483
#>  [2,]  1.0488088  0.66011331 -0.1097643 -0.5218446  0.43838004  0.4082483
#>  [3,]  1.0488088 -1.14802316 -1.2622890 -0.9566819  0.34288231  0.4082483
#>  [4,] -0.9534626 -1.14802316 -0.1097643  0.2976565 -1.12778267 -1.0206207
#>  [5,] -0.9534626  0.66011331 -0.1097643 -0.4582915  0.47657913  0.4082483
#>  [6,] -0.9534626  0.66011331 -0.1097643 -0.4582915  0.47657913  0.4082483
#>  [7,] -0.9534626  0.05740116  1.0427605  0.4465047 -1.14688222 -1.0206207
#>  [8,] -0.9534626  0.05740116  1.0427605  0.4465047 -1.14688222 -1.0206207
#>  [9,] -0.9534626  0.66011331  1.0427605  2.0871793 -1.41427585 -1.0206207
#> [10,] -0.9534626  0.66011331  1.0427605  1.8195871 -0.84128950 -1.0206207
#> [11,]  1.0488088 -1.14802316 -1.2622890 -1.2016960  0.78217186  0.4082483
#> [12,]  1.0488088 -0.54531100 -1.2622890 -1.2267828  2.40563321  0.4082483
#> [13,]  1.0488088 -1.14802316 -1.2622890 -1.2652492  1.04956549  0.4082483
#> [14,] -0.9534626 -1.14802316 -1.2622890 -0.8554986  0.05638913 -1.0206207
#> [15,] -0.9534626 -0.54531100  1.0427605  0.7993919 -1.73896812 -1.0206207
#> [16,] -0.9534626 -0.54531100  1.0427605  0.6823203 -0.99408586 -1.0206207
#> [17,] -0.9534626  0.66011331  1.0427605  1.0669841  0.11368777 -1.0206207
#> [18,]  1.0488088 -0.54531100 -1.2622890 -1.0645550  0.19008595  1.8371173
#> [19,]  1.0488088  0.66011331  1.0427605  1.0753463  1.04956549  1.8371173
#> [20,]  1.0488088  3.07096194  1.0427605  0.6572335 -0.24920359  1.8371173
#> [21,]  1.0488088 -0.54531100 -1.2622890 -0.8479726  0.83947049  0.4082483
#>                hp        qsec         vs          wt
#>  [1,] -0.53457475 -0.79426584 -0.9534626 -0.53757893
#>  [2,] -0.53457475 -0.42868595 -0.9534626 -0.28118998
#>  [3,] -0.77312873  0.60929982  1.0488088 -0.83921300
#>  [4,] -0.53457475  1.15114145  1.0488088  0.06066196
#>  [5,] -0.35215112  0.40692524  1.0488088  0.28688751
#>  [6,] -0.35215112  0.79861798  1.0488088  0.28688751
#>  [7,]  0.44770635 -0.18061388 -0.9534626  0.92031904
#>  [8,]  0.44770635 -0.05004963 -0.9534626  0.57846711
#>  [9,]  0.79852103  0.19802244 -0.9534626  2.10674636
#> [10,]  1.14933571 -0.16755745 -0.9534626  2.20226382
#> [11,] -1.15200859  1.17072609  1.0488088 -0.95986662
#> [12,] -1.34846481  0.55054591  1.0488088 -1.54805305
#> [13,] -1.16604117  1.45143922  1.0488088 -1.32685473
#> [14,] -0.71699838  1.52324955  1.0488088 -0.69342320
#> [15,]  0.02672874 -0.52660913 -0.9534626  0.36732326
#> [16,]  0.02672874 -0.24589600 -0.9534626  0.28186028
#> [17,]  1.35982452 -1.47972814 -0.9534626  0.68906626
#> [18,] -0.49247699 -0.50702449  1.0488088 -1.65060863
#> [19,]  1.62644368 -2.07379546 -0.9534626  0.01541685
#> [20,]  2.62275737 -2.00851334 -0.9534626  0.41759560
#> [21,] -0.54860734  0.60277161  1.0488088 -0.37670743
#> attr(,"center")
#>  [1]   0.4761905   2.9047619   6.1904762 222.4047619   3.6704762   3.7142857
#>  [7] 148.0952381  17.6766667   0.4761905   3.1546667
#> attr(,"scale")
#>  [1]   0.4994328   1.6591668   1.7353207 119.5849508   0.5235727   0.6998542
#>  [7]  71.2626963   1.5318129   0.4994328   0.9945826
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 43
#> 
#> $lambda.min
#> [1] 0.2939522
#> 
#> $null.dev
#> [1] 40.66181
#> 
#> $Bias
#> [1] 4.283259
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 6.284955