Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 27.999443 26.660223 23.972808 21.057552 17.486126 14.529215 12.096713
#>   [8] 10.110930  8.504926  7.221071  6.209791  5.428483  4.840571  4.414689
#>  [15]  4.123974  3.945148  3.783082  3.683819  3.669101  3.670379  3.670379
#>  [22]  3.670379  3.670379  3.670379  3.670379  3.670379  3.670379  3.670379
#>  [29]  3.670379  3.674795  3.731909  3.806395  3.884336  3.964298  4.041194
#>  [36]  4.152596  4.481138  4.658724  4.814238  4.734213  4.738484  4.708446
#>  [43]  4.616939  4.672794  4.222204  6.400152  6.652139  6.655736  6.655243
#>  [50]  6.800779  6.781970  7.055290  6.639279  6.842905  8.240176 12.676979
#>  [57] 12.751283 12.746687 12.686384 12.683394 12.542039 12.540688 13.754023
#>  [64] 13.898722 13.899551 13.957808 13.803584 26.210307 26.658731 26.931562
#>  [71] 26.775967 26.786139 27.036781 27.038666 27.040533 27.042382 27.044213
#>  [78] 28.044071 28.124210 28.153336 28.097637 28.030396 28.033681 28.036909
#>  [85] 28.040080 28.043194 28.046254 28.049258 28.112053 28.114775 28.231237
#>  [92] 28.233780 28.236278 28.238733 28.241146 28.243517 28.237373 28.239614
#>  [99] 28.241818 28.244047
#> 
#> $cvse
#>   [1]  9.2236451  9.1378772  8.4575624  7.6299507  6.4000686  5.3801306
#>   [7]  4.5355631  3.8361040  3.2554053  2.7707838  2.3630822  2.0165863
#>  [13]  1.7189638  1.4612221  1.2377404  1.0464732  0.8981115  0.7913363
#>  [19]  0.7413879  0.7291831  0.7291831  0.7291831  0.7291831  0.7291831
#>  [25]  0.7291831  0.7291831  0.7291831  0.7291831  0.7291831  0.7284411
#>  [31]  0.7362627  0.7443063  0.7537862  0.7648385  0.7761648  0.8133012
#>  [37]  0.8720016  0.9357497  0.9888595  1.0204338  1.0202113  1.0202367
#>  [43]  1.0260516  1.0119429  1.0206363  1.9152792  2.0569156  2.0541875
#>  [49]  2.0499229  2.0346868  2.0323715  2.0427916  1.8577906  1.8601080
#>  [55]  2.0280857  4.0265952  4.0191167  4.0192322  4.0162467  4.0164813
#>  [61]  4.0302484  4.0304867  4.3953894  4.3971636  4.3976654  4.3956752
#>  [67]  4.4175777 13.3351898 13.3780593 13.3709827 13.3791630 13.3794938
#>  [73] 13.3851758 13.3863688 13.3875529 13.3887285 13.3898956 13.5466796
#>  [79] 13.5458906 13.5462260 13.5535051 13.5586632 13.5600350 13.5613931
#>  [85] 13.5627376 13.5640687 13.5653867 13.5666919 13.5649437 13.5661915
#>  [91] 13.5940614 13.5953417 13.5966098 13.5978660 13.5991103 13.6003431
#>  [97] 13.6024187 13.6036337 13.6048375 13.6060272
#> 
#> $fold
#>  [1]  6  1  3 10  8  3  9  9  8  2  5  2  7  7 10  4  1  5  1  6  4
#> 
#> $lambda
#>   [1] 4.866025058 4.538071255 4.232220441 3.946982948 3.680969507 3.432884482
#>   [7] 3.201519558 2.985747855 2.784518442 2.596851219 2.421832138 2.258608757
#>  [13] 2.106386086 1.964422714 1.832027198 1.708554696 1.593403828 1.486013743
#>  [19] 1.385861390 1.292458971 1.205351562 1.124114901 1.048353319 0.977697814
#>  [25] 0.911804253 0.850351698 0.793040840 0.739592542 0.689746480 0.643259875
#>  [31] 0.599906313 0.559474635 0.521767917 0.486602506 0.453807126 0.423222044
#>  [37] 0.394698295 0.368096951 0.343288448 0.320151956 0.298574785 0.278451844
#>  [43] 0.259685122 0.242183214 0.225860876 0.210638609 0.196442271 0.183202720
#>  [49] 0.170855470 0.159340383 0.148601374 0.138586139 0.129245897 0.120535156
#>  [55] 0.112411490 0.104835332 0.097769781 0.091180425 0.085035170 0.079304084
#>  [61] 0.073959254 0.068974648 0.064325988 0.059990631 0.055947464 0.052176792
#>  [67] 0.048660251 0.045380713 0.042322204 0.039469829 0.036809695 0.034328845
#>  [73] 0.032015196 0.029857479 0.027845184 0.025968512 0.024218321 0.022586088
#>  [79] 0.021063861 0.019644227 0.018320272 0.017085547 0.015934038 0.014860137
#>  [85] 0.013858614 0.012924590 0.012053516 0.011241149 0.010483533 0.009776978
#>  [91] 0.009118043 0.008503517 0.007930408 0.007395925 0.006897465 0.006432599
#>  [97] 0.005999063 0.005594746 0.005217679 0.004866025
#> 
#> $fit
#> $beta
#>               4.8660    4.5381     4.2322   3.9470    3.6810   3.4329    3.2015
#> (Intercept) 19.20952 20.867214 22.4131814 23.85496 25.199559 26.45354 27.623009
#> am           0.00000  0.000000  0.0000000  0.00000  0.000000  0.00000  0.000000
#> carb         0.00000  0.000000  0.0000000  0.00000  0.000000  0.00000  0.000000
#> cyl          0.00000  0.000000  0.0000000  0.00000  0.000000  0.00000  0.000000
#> disp         0.00000  0.000000  0.0000000  0.00000  0.000000  0.00000  0.000000
#> drat         0.00000  0.000000  0.0000000  0.00000  0.000000  0.00000  0.000000
#> gear         0.00000  0.000000  0.0000000  0.00000  0.000000  0.00000  0.000000
#> hp           0.00000  0.000000  0.0000000  0.00000  0.000000  0.00000  0.000000
#> qsec         0.00000  0.000000  0.0000000  0.00000  0.000000  0.00000  0.000000
#> vs           0.00000  0.000000  0.0000000  0.00000  0.000000  0.00000  0.000000
#> wt           0.00000 -0.511242 -0.9880281 -1.43268 -1.847365 -2.23410 -2.594772
#>                2.9857    2.7845   2.5969    2.4218   2.2586    2.1064    1.9644
#> (Intercept) 28.713658 29.730802 30.67939 31.564052 32.38909 33.158520 33.876095
#> am           0.000000  0.000000  0.00000  0.000000  0.00000  0.000000  0.000000
#> carb         0.000000  0.000000  0.00000  0.000000  0.00000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.00000  0.000000  0.00000  0.000000  0.000000
#> disp         0.000000  0.000000  0.00000  0.000000  0.00000  0.000000  0.000000
#> drat         0.000000  0.000000  0.00000  0.000000  0.00000  0.000000  0.000000
#> gear         0.000000  0.000000  0.00000  0.000000  0.00000  0.000000  0.000000
#> hp           0.000000  0.000000  0.00000  0.000000  0.00000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.00000  0.000000  0.00000  0.000000  0.000000
#> vs           0.000000  0.000000  0.00000  0.000000  0.00000  0.000000  0.000000
#> wt          -2.931135 -3.244828 -3.53738 -3.810214 -4.06466 -4.301958 -4.523262
#>                1.8320    1.7086    1.5934    1.4860    1.3859    1.2925
#> (Intercept) 34.545307 35.169417 35.606878 35.606878 35.606878 35.606878
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -4.729652 -4.922131 -5.057047 -5.057047 -5.057047 -5.057047
#>                1.2054    1.1241    1.0484    0.9777    0.9118    0.8504
#> (Intercept) 35.606878 35.606878 35.606878 35.606878 35.606878 35.606878
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -5.057047 -5.057047 -5.057047 -5.057047 -5.057047 -5.057047
#>                0.7930    0.7396    0.6897    0.6433    0.5999    0.5595
#> (Intercept) 35.606878 35.606878 35.606878 35.606878 35.606878 35.606878
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -5.057047 -5.057047 -5.057047 -5.057047 -5.057047 -5.057047
#>                0.5218     0.4866     0.4538       0.4232      0.3947
#> (Intercept) 35.606878 35.4351605 35.2686650 35.292598137 35.40859508
#> am           0.000000  0.0000000  0.0000000  0.000000000  0.00000000
#> carb         0.000000  0.0000000  0.0000000  0.000000000  0.00000000
#> cyl          0.000000  0.0000000  0.0000000  0.000000000  0.00000000
#> disp         0.000000  0.0000000  0.0000000  0.000000000  0.00000000
#> drat         0.000000  0.0000000  0.0000000  0.000000000  0.00000000
#> gear         0.000000  0.0000000  0.0000000  0.000000000  0.00000000
#> hp           0.000000  0.0000000  0.0000000 -0.001656157 -0.00407926
#> qsec         0.000000  0.0000000  0.0000000  0.000000000  0.00000000
#> vs           0.000000  0.1524087  0.3001826  0.306673355  0.24429506
#> wt          -5.057047 -5.0219943 -4.9880078 -4.921433275 -4.84055938
#>                   0.3681       0.3433      0.3202      0.2986      0.2785
#> (Intercept) 35.516462941 35.360265174 34.09507523 32.93613632 31.85694425
#> am           0.000000000  0.000000000  0.00000000  0.00000000  0.00000000
#> carb         0.000000000  0.000000000  0.00000000  0.00000000  0.00000000
#> cyl          0.000000000  0.000000000  0.00000000  0.00000000  0.00000000
#> disp         0.000000000  0.000000000  0.00000000  0.00000000  0.00000000
#> drat         0.000000000  0.050253638  0.31155874  0.55161593  0.77531489
#> gear         0.000000000  0.000000000  0.00000000  0.00000000  0.00000000
#> hp          -0.006335935 -0.008301365 -0.00934502 -0.01034473 -0.01128482
#> qsec         0.000000000  0.000000000  0.00000000  0.00000000  0.00000000
#> vs           0.186344875  0.141529511  0.15691370  0.16873873  0.17921405
#> wt          -4.765207185 -4.677357959 -4.52402317 -4.38207953 -4.24959939
#>                  0.2597      0.2422      0.2259      0.2106       0.1964
#> (Intercept) 30.84922400 29.80718328 29.25946007 25.11034025 20.153841036
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.000000000
#> carb         0.00000000  0.00000000  0.00000000 -0.50309609 -0.907060324
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.000000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.000000000
#> drat         0.98405507  1.18682580  1.27255858  2.26801033  2.005427070
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  1.317186372
#> hp          -0.01215448 -0.01237263 -0.01189381  0.00000000  0.000000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.000000000
#> vs           0.18947800  0.24499445  0.35399892  0.09992553  0.009401855
#> wt          -4.12616323 -4.02040662 -3.97846111 -3.81492121 -3.107904381
#>                  0.1832     0.1709     0.1593     0.1486     0.1386     0.1292
#> (Intercept) 20.14755992 20.1635242 20.1801337 20.1957620 20.2077732 20.2213183
#> am           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb        -0.89648948 -0.8846001 -0.8733464 -0.8628378 -0.8532502 -0.8441172
#> cyl          0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         1.98572596  1.9678059  1.9511448  1.9356107  1.9211476  1.9076288
#> gear         1.31557106  1.3088819  1.3022187  1.2959704  1.2906726  1.2852567
#> hp           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> vs           0.09570868  0.1765113  0.2518949  0.3222000  0.3877345  0.4488811
#> wt          -3.10278939 -3.1012937 -3.1001608 -3.0991253 -3.0977929 -3.0968835
#>                 0.1205     0.1124       0.1048       0.0978       0.0912
#> (Intercept) 20.2313296 20.2432125 19.433582760 19.435074123 19.436402944
#> am           0.0000000  0.0000000  0.000000000  0.000000000  0.000000000
#> carb        -0.8358050 -0.8278548 -0.738643758 -0.738503490 -0.738378522
#> cyl          0.0000000  0.0000000  0.000000000  0.000000000  0.000000000
#> disp         0.0000000  0.0000000  0.009086912  0.009089507  0.009091819
#> drat         1.8950861  1.8833218  2.185549909  2.185570439  2.185588696
#> gear         1.2807069  1.2759808  1.208987428  1.208675389  1.208397393
#> hp           0.0000000  0.0000000  0.000000000  0.000000000  0.000000000
#> qsec         0.0000000  0.0000000  0.000000000  0.000000000  0.000000000
#> vs           0.5058715  0.5590550  1.079042014  1.079167406  1.079279121
#> wt          -3.0956697 -3.0948920 -3.885932983 -3.886391374 -3.886799780
#>                   0.0850       0.0793       0.0740       0.0690       0.0643
#> (Intercept) 19.437586932 19.438641865 19.439581801 19.440419269 21.112877755
#> am           0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> carb        -0.738267184 -0.738167988 -0.738079610 -0.738000870 -0.506200304
#> cyl          0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> disp         0.009093878  0.009095713  0.009097348  0.009098805  0.012707417
#> drat         2.185604940  2.185619396  2.185632266  2.185643725  1.830812002
#> gear         1.208149719  1.207929058  1.207732461  1.207557302  1.256926386
#> hp           0.000000000  0.000000000  0.000000000  0.000000000 -0.009310475
#> qsec         0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> vs           1.079378650  1.079467321  1.079546319  1.079616699  1.153091915
#> wt          -3.887163650 -3.887487840 -3.887776675 -3.888034011 -4.136349286
#>                   0.0600       0.0559       0.0522       0.0487      0.0454
#> (Intercept) 21.114678270 21.116347203 21.117894166 21.119328071 21.42257024
#> am           0.000000000  0.000000000  0.000000000  0.000000000  0.52537507
#> carb        -0.506013206 -0.505839781 -0.505679030 -0.505530026 -0.52221793
#> cyl          0.000000000  0.000000000  0.000000000  0.000000000  0.00000000
#> disp         0.012710802  0.012713940  0.012716848  0.012719544  0.01218034
#> drat         1.830672500  1.830543219  1.830423407  1.830312368  1.76645157
#> gear         1.256714833  1.256518715  1.256336911  1.256168378  1.07297678
#> hp          -0.009313373 -0.009316058 -0.009318548 -0.009320855 -0.00876748
#> qsec         0.000000000  0.000000000  0.000000000  0.000000000  0.00000000
#> vs           1.153244575  1.153386078  1.153517239  1.153638815  1.27393787
#> wt          -4.136815875 -4.137248382 -4.137649292 -4.138020914 -4.00245424
#>                   0.0423       0.0395       0.0368       0.0343       0.0320
#> (Intercept) 21.420599612 21.418772170 21.417078094 21.415508148 17.963727592
#> am           0.525330064  0.525287117  0.525246269  0.525207531  0.847159436
#> carb        -0.522418105 -0.522603670 -0.522775641 -0.522934965 -0.671443943
#> cyl          0.000000000  0.000000000  0.000000000  0.000000000  0.316716202
#> disp         0.012176766  0.012173455  0.012170393  0.012167561  0.010459515
#> drat         1.766613338  1.766762401  1.766899777  1.767026396  1.962226839
#> gear         1.073216013  1.073439315  1.073647558  1.073841600  1.278247709
#> hp          -0.008764359 -0.008761479 -0.008758821 -0.008756368 -0.009074509
#> qsec         0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> vs           1.273763109  1.273600996  1.273450670  1.273311322  1.579753812
#> wt          -4.001966836 -4.001515310 -4.001097131 -4.000709929 -3.801884846
#>                   0.0299       0.0278       0.0260       0.0242       0.0226
#> (Intercept) 17.959732000 17.955947209 17.952361456 17.948963735 17.945743737
#> am           0.847566870  0.847951109  0.848313642  0.848655845  0.848978995
#> carb        -0.671665947 -0.671876248 -0.672075488 -0.672264272 -0.672443167
#> cyl          0.317012894  0.317293723  0.317559610  0.317811409  0.318049917
#> disp         0.010457315  0.010455227  0.010453245  0.010451364  0.010449578
#> drat         1.962438091  1.962638466  1.962828524  1.963008792  1.963179772
#> gear         1.278532561  1.278802785  1.279059167  1.279302442  1.279533304
#> hp          -0.009073797 -0.009073122 -0.009072481 -0.009071873 -0.009071297
#> qsec         0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> vs           1.580042168  1.580314391  1.580571514  1.580814487  1.581044190
#> wt          -3.801538637 -3.801210425 -3.800899276 -3.800604305 -3.800324674
#>                   0.0211       0.0196       0.0183       0.0171       0.0159
#> (Intercept) 17.942691790 17.939798818 17.937056290 17.934456183 17.931990943
#> am           0.849284271  0.849572767  0.849845499  0.850103410  0.850347378
#> carb        -0.672612708 -0.672773397 -0.672925705 -0.673070078 -0.673206936
#> cyl          0.318275881  0.318489997  0.318692916  0.318885251  0.319067573
#> disp         0.010447885  0.010446278  0.010444754  0.010443309  0.010441938
#> drat         1.963341938  1.963495740  1.963641605  1.963779936  1.963911119
#> gear         1.279752399  1.279960340  1.280157702  1.280345026  1.280522824
#> hp          -0.009070751 -0.009070235 -0.009069746 -0.009069284 -0.009068846
#> qsec         0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> vs           1.581261433  1.581466969  1.581661494  1.581845653  1.582020049
#> wt          -3.800059591 -3.799808302 -3.799570094 -3.799344292 -3.799130255
#>                   0.0149       0.0139       0.0129       0.0121       0.0112
#> (Intercept) 17.929653453 17.927437001 17.925335256 17.923342236 23.083529287
#> am           0.850578219  0.850796693  0.851003511  0.851199334  0.801296084
#> carb        -0.673336676 -0.673459671 -0.673576275 -0.673686820 -0.698179247
#> cyl          0.319240421  0.319404302  0.319559691  0.319707037  0.228215768
#> disp         0.010440639  0.010439408  0.010438240  0.010437134  0.009471336
#> drat         1.964035517  1.964153476  1.964265326  1.964371378  1.982397169
#> gear         1.280691581  1.280851754  1.281003775  1.281148056  1.046403278
#> hp          -0.009068433 -0.009068042 -0.009067672 -0.009067323 -0.009273256
#> qsec         0.000000000  0.000000000  0.000000000  0.000000000 -0.237020850
#> vs           1.582185239  1.582341744  1.582490051  1.582630613  1.861215019
#> wt          -3.798927374 -3.798735073 -3.798552805 -3.798380050 -3.597970019
#>                   0.0105       0.0098       0.0091       0.0085       0.0079
#> (Intercept) 23.090669192 23.097585226 23.104284633 23.110774411 23.117061319
#> am           0.801303302  0.801311225  0.801319783  0.801328908  0.801338538
#> carb        -0.698274160 -0.698366728 -0.698456994 -0.698545002 -0.698630795
#> cyl          0.228126662  0.228040925  0.227958417  0.227879006  0.227802565
#> disp         0.009468306  0.009465362  0.009462502  0.009459724  0.009457025
#> drat         1.982508667  1.982617292  1.982723105  1.982826167  1.982926536
#> gear         1.046061486  1.045730643  1.045410384  1.045100355  1.044800216
#> hp          -0.009272993 -0.009272737 -0.009272488 -0.009272245 -0.009272007
#> qsec        -0.237388845 -0.237745777 -0.238091980 -0.238427778 -0.238753483
#> vs           1.861656098  1.862084592  1.862500844  1.862905187  1.863297949
#> wt          -3.597485688 -3.597015143 -3.596558010 -3.596113924 -3.595682529
#>                   0.0074       0.0069       0.0064       0.0060       0.0056
#> (Intercept) 23.123151891 23.129052438 23.134769060 23.140307650 23.145673907
#> am           0.801348617  0.801359090  0.801369909  0.801381026  0.801392399
#> carb        -0.698714417 -0.698795911 -0.698875321 -0.698952690 -0.699028061
#> cyl          0.227728973  0.227658115  0.227589880  0.227524162  0.227460861
#> disp         0.009454403  0.009451856  0.009449383  0.009446980  0.009444647
#> drat         1.983024273  1.983119436  1.983212083  1.983302272  1.983390059
#> gear         1.044509638  1.044228304  1.043955907  1.043692153  1.043436758
#> hp          -0.009271776 -0.009271551 -0.009271331 -0.009271118 -0.009270909
#> qsec        -0.239069400 -0.239375823 -0.239673038 -0.239961324 -0.240240947
#> vs           1.863679446  1.864049988  1.864409879  1.864759412  1.865098874
#> wt          -3.595263478 -3.594856434 -3.594461065 -3.594077049 -3.593704075
#>                   0.0052       0.0049
#> (Intercept) 23.150873338 23.155911265
#> am           0.801403988  0.801415757
#> carb        -0.699101476 -0.699172977
#> cyl          0.227399880  0.227341126
#> disp         0.009442381  0.009440180
#> drat         1.983475500  1.983558649
#> gear         1.043189445  1.042949950
#> hp          -0.009270706 -0.009270509
#> qsec        -0.240512170 -0.240775245
#> vs           1.865428546  1.865748700
#> wt          -3.593341834 -3.592990031
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1  10   9  18
#>  [37]  24  23  35  25  24  24  23  17   8  37  51  13  13  13  13  12  12  11
#>  [55]  11  62   1   1   1   1   1   1  70   1   1   1   1  63   1   1   1   1
#>  [73]  71   1   1   1   1   1   1   1   1   1   1   1   1   1   1 101   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.866025058 4.538071255 4.232220441 3.946982948 3.680969507 3.432884482
#>   [7] 3.201519558 2.985747855 2.784518442 2.596851219 2.421832138 2.258608757
#>  [13] 2.106386086 1.964422714 1.832027198 1.708554696 1.593403828 1.486013743
#>  [19] 1.385861390 1.292458971 1.205351562 1.124114901 1.048353319 0.977697814
#>  [25] 0.911804253 0.850351698 0.793040840 0.739592542 0.689746480 0.643259875
#>  [31] 0.599906313 0.559474635 0.521767917 0.486602506 0.453807126 0.423222044
#>  [37] 0.394698295 0.368096951 0.343288448 0.320151956 0.298574785 0.278451844
#>  [43] 0.259685122 0.242183214 0.225860876 0.210638609 0.196442271 0.183202720
#>  [49] 0.170855470 0.159340383 0.148601374 0.138586139 0.129245897 0.120535156
#>  [55] 0.112411490 0.104835332 0.097769781 0.091180425 0.085035170 0.079304084
#>  [61] 0.073959254 0.068974648 0.064325988 0.059990631 0.055947464 0.052176792
#>  [67] 0.048660251 0.045380713 0.042322204 0.039469829 0.036809695 0.034328845
#>  [73] 0.032015196 0.029857479 0.027845184 0.025968512 0.024218321 0.022586088
#>  [79] 0.021063861 0.019644227 0.018320272 0.017085547 0.015934038 0.014860137
#>  [85] 0.013858614 0.012924590 0.012053516 0.011241149 0.010483533 0.009776978
#>  [91] 0.009118043 0.008503517 0.007930408 0.007395925 0.006897465 0.006432599
#>  [97] 0.005999063 0.005594746 0.005217679 0.004866025
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 35
#> 
#> $loss
#>   [1] 560.01810 464.56263 384.69994 318.18610 263.08382 217.72152 180.65783
#>   [8] 150.65076 126.63095 107.67845  93.00258  81.92444  73.86178  68.31585
#>  [15]  64.86000  63.12981  62.77590  62.77590  62.77590  62.77590  62.77590
#>  [22]  62.77590  62.77590  62.77590  62.77590  62.77590  62.77590  62.77590
#>  [29]  62.77590  62.77590  62.77590  62.77590  62.77590  61.24857  59.94175
#>  [36]  58.15521  56.49025  55.33270  54.19741  51.87279  50.23353  49.13819
#>  [43]  48.48991  48.00340  47.56787  44.26244  40.22041  39.90124  39.64576
#>  [50]  39.44518  39.29098  39.17582  39.09321  39.03784  39.00495  36.53502
#>  [57]  36.53501  36.53500  36.53500  36.53499  36.53499  36.53499  35.73504
#>  [64]  35.73503  35.73502  35.73502  35.73501  35.48672  35.48670  35.48669
#>  [71]  35.48668  35.48668  35.13181  35.13180  35.13178  35.13177  35.13176
#>  [78]  35.13175  35.13174  35.13174  35.13173  35.13173  35.13172  35.13172
#>  [85]  35.13171  35.13171  35.13170  35.04751  35.04750  35.04748  35.04747
#>  [92]  35.04746  35.04744  35.04743  35.04742  35.04741  35.04740  35.04739
#>  [99]  35.04738  35.04738
#> 
#> $linear.predictors
#>      4.8660   4.5381   4.2322   3.9470   3.6810   3.4329   3.2015   2.9857
#> 1  19.20952 19.52776 19.82455 20.10133 20.35946 20.60020 20.82471 21.03409
#> 2  19.20952 19.39739 19.57260 19.73600 19.88839 20.03050 20.16304 20.28665
#> 3  19.20952 19.68113 20.12096 20.53114 20.91367 21.27043 21.60314 21.91343
#> 4  19.20952 19.22357 19.23667 19.24889 19.26028 19.27091 19.28082 19.29006
#> 5  19.20952 19.10854 19.01436 18.92654 18.84463 18.76824 18.69699 18.63055
#> 6  19.20952 19.09832 18.99460 18.89788 18.80768 18.72355 18.64510 18.57193
#> 7  19.20952 19.10854 19.01436 18.92654 18.84463 18.76824 18.69699 18.63055
#> 8  19.20952 19.10854 19.01436 18.92654 18.84463 18.76824 18.69699 18.63055
#> 9  19.20952 18.78646 18.39191 18.02395 17.68079 17.36075 17.06229 16.78394
#> 10 19.20952 18.96028 18.72784 18.51106 18.30889 18.12035 17.94451 17.78053
#> 11 19.20952 18.93472 18.67844 18.43942 18.21652 18.00864 17.81477 17.63397
#> 12 19.20952 18.18319 17.22603 16.33338 15.50090 14.72451 14.00046 13.32520
#> 13 19.20952 18.09424 17.05412 16.08410 15.17945 14.33578 13.54897 12.81518
#> 14 19.20952 20.04156 20.81752 21.54118 22.21607 22.84547 23.43245 23.97988
#> 15 19.20952 19.60700 19.97769 20.32340 20.64581 20.94648 21.22690 21.48841
#> 16 19.20952 19.06764 18.93532 18.81192 18.69684 18.58951 18.48941 18.39606
#> 17 19.20952 19.11110 19.01930 18.93370 18.85386 18.77941 18.70997 18.64521
#> 18 19.20952 19.77316 20.29880 20.78902 21.24620 21.67257 22.07020 22.44103
#> 19 19.20952 20.09370 20.91829 21.68731 22.40450 23.07335 23.69712 24.27885
#> 20 19.20952 19.45107 19.67634 19.88643 20.08236 20.26508 20.43549 20.59442
#> 21 19.20952 19.04208 18.88592 18.74029 18.60447 18.47780 18.35967 18.24951
#>      2.7845   2.5969   2.4218   2.2586   2.1064    1.9644    1.8320    1.7086
#> 1  21.22935 21.41146 21.58129 21.73968 21.88739 22.025147 22.153619 22.273433
#> 2  20.40192 20.50943 20.60969 20.70319 20.79039 20.871715 20.947558 21.018290
#> 3  22.20280 22.47267 22.72436 22.95908 23.17798 23.382126 23.572515 23.750073
#> 4  19.29868 19.30672 19.31421 19.32121 19.32773 19.333806 19.339477 19.344765
#> 5  18.56859 18.51081 18.45692 18.40666 18.35978 18.316072 18.275305 18.237286
#> 6  18.50370 18.44006 18.38071 18.32536 18.27375 18.225607 18.180712 18.138843
#> 7  18.56859 18.51081 18.45692 18.40666 18.35978 18.316072 18.275305 18.237286
#> 8  18.56859 18.51081 18.45692 18.40666 18.35978 18.316072 18.275305 18.237286
#> 9  16.52435 16.28226 16.05648 15.84592 15.64955 15.466416 15.295624 15.136343
#> 10 17.62759 17.48497 17.35195 17.22791 17.11222 17.004326 16.903706 16.809868
#> 11 17.46535 17.30810 17.16144 17.02467 16.89712 16.778163 16.667223 16.563761
#> 12 12.69545 12.10815 11.56043 11.04962 10.57324 10.128967  9.714635  9.328228
#> 13 12.13085 11.49265 10.89745 10.34237  9.82470  9.341919  8.891676  8.471778
#> 14 24.49040 24.96652 25.41056 25.82466 26.21086 26.571026 26.906919 27.220175
#> 15 21.73230 21.95975 22.17187 22.36970 22.55419 22.726253 22.886715 23.036364
#> 16 18.30901 18.22782 18.15210 18.08148 18.01563 17.954211 17.896933 17.843515
#> 17 18.58482 18.52849 18.47597 18.42698 18.38129 18.338688 18.298953 18.261896
#> 18 22.78687 23.10940 23.41019 23.69072 23.95233 24.196313 24.423852 24.636056
#> 19 24.82138 25.32734 25.79920 26.23926 26.64966 27.032399 27.389344 27.722232
#> 20 20.74263 20.88085 21.00976 21.12998 21.24210 21.346658 21.444172 21.535114
#> 21 18.14677 18.05095 17.96159 17.87825 17.80053 17.728048 17.660450 17.597409
#>       1.5934    1.4860    1.3859    1.2925    1.2054    1.1241    1.0484
#> 1  22.357415 22.357415 22.357415 22.357415 22.357415 22.357415 22.357415
#> 2  21.067868 21.067868 21.067868 21.067868 21.067868 21.067868 21.067868
#> 3  23.874529 23.874529 23.874529 23.874529 23.874529 23.874529 23.874529
#> 4  19.348472 19.348472 19.348472 19.348472 19.348472 19.348472 19.348472
#> 5  18.210637 18.210637 18.210637 18.210637 18.210637 18.210637 18.210637
#> 6  18.109496 18.109496 18.109496 18.109496 18.109496 18.109496 18.109496
#> 7  18.210637 18.210637 18.210637 18.210637 18.210637 18.210637 18.210637
#> 8  18.210637 18.210637 18.210637 18.210637 18.210637 18.210637 18.210637
#> 9  15.024697 15.024697 15.024697 15.024697 15.024697 15.024697 15.024697
#> 10 16.744093 16.744093 16.744093 16.744093 16.744093 16.744093 16.744093
#> 11 16.491241 16.491241 16.491241 16.491241 16.491241 16.491241 16.491241
#> 12  9.057382  9.057382  9.057382  9.057382  9.057382  9.057382  9.057382
#> 13  8.177456  8.177456  8.177456  8.177456  8.177456  8.177456  8.177456
#> 14 27.439747 27.439747 27.439747 27.439747 27.439747 27.439747 27.439747
#> 15 23.141257 23.141257 23.141257 23.141257 23.141257 23.141257 23.141257
#> 16 17.806073 17.806073 17.806073 17.806073 17.806073 17.806073 17.806073
#> 17 18.235922 18.235922 18.235922 18.235922 18.235922 18.235922 18.235922
#> 18 24.784798 24.784798 24.784798 24.784798 24.784798 24.784798 24.784798
#> 19 27.955566 27.955566 27.955566 27.955566 27.955566 27.955566 27.955566
#> 20 21.598858 21.598858 21.598858 21.598858 21.598858 21.598858 21.598858
#> 21 17.553221 17.553221 17.553221 17.553221 17.553221 17.553221 17.553221
#>       0.9777    0.9118    0.8504    0.7930    0.7396    0.6897    0.6433
#> 1  22.357415 22.357415 22.357415 22.357415 22.357415 22.357415 22.357415
#> 2  21.067868 21.067868 21.067868 21.067868 21.067868 21.067868 21.067868
#> 3  23.874529 23.874529 23.874529 23.874529 23.874529 23.874529 23.874529
#> 4  19.348472 19.348472 19.348472 19.348472 19.348472 19.348472 19.348472
#> 5  18.210637 18.210637 18.210637 18.210637 18.210637 18.210637 18.210637
#> 6  18.109496 18.109496 18.109496 18.109496 18.109496 18.109496 18.109496
#> 7  18.210637 18.210637 18.210637 18.210637 18.210637 18.210637 18.210637
#> 8  18.210637 18.210637 18.210637 18.210637 18.210637 18.210637 18.210637
#> 9  15.024697 15.024697 15.024697 15.024697 15.024697 15.024697 15.024697
#> 10 16.744093 16.744093 16.744093 16.744093 16.744093 16.744093 16.744093
#> 11 16.491241 16.491241 16.491241 16.491241 16.491241 16.491241 16.491241
#> 12  9.057382  9.057382  9.057382  9.057382  9.057382  9.057382  9.057382
#> 13  8.177456  8.177456  8.177456  8.177456  8.177456  8.177456  8.177456
#> 14 27.439747 27.439747 27.439747 27.439747 27.439747 27.439747 27.439747
#> 15 23.141257 23.141257 23.141257 23.141257 23.141257 23.141257 23.141257
#> 16 17.806073 17.806073 17.806073 17.806073 17.806073 17.806073 17.806073
#> 17 18.235922 18.235922 18.235922 18.235922 18.235922 18.235922 18.235922
#> 18 24.784798 24.784798 24.784798 24.784798 24.784798 24.784798 24.784798
#> 19 27.955566 27.955566 27.955566 27.955566 27.955566 27.955566 27.955566
#> 20 21.598858 21.598858 21.598858 21.598858 21.598858 21.598858 21.598858
#> 21 17.553221 17.553221 17.553221 17.553221 17.553221 17.553221 17.553221
#>       0.5999    0.5595    0.5218    0.4866    0.4538    0.4232   0.3947
#> 1  22.357415 22.357415 22.357415 22.277535 22.200085 22.216266 22.27761
#> 2  21.067868 21.067868 21.067868 20.996927 20.928143 20.961300 21.04327
#> 3  23.874529 23.874529 23.874529 23.936542 23.996670 24.027524 24.04342
#> 4  19.348472 19.348472 19.348472 19.441857 19.532403 19.594686 19.64177
#> 5  18.210637 18.210637 18.210637 18.159500 18.109918 18.073040 18.04320
#> 6  18.109496 18.109496 18.109496 18.211469 18.310341 18.397216 18.47623
#> 7  18.210637 18.210637 18.210637 18.311909 18.410101 18.465834 18.49962
#> 8  18.210637 18.210637 18.210637 18.311909 18.410101 18.465834 18.49962
#> 9  15.024697 15.024697 15.024697 14.995643 14.967473 14.964257 14.97325
#> 10 16.744093 16.744093 16.744093 16.703122 16.663396 16.637544 16.61904
#> 11 16.491241 16.491241 16.491241 16.452022 16.413996 16.391472 16.37701
#> 12  9.057382  9.057382  9.057382  9.069690  9.081624  9.115561  9.15941
#> 13  8.177456  8.177456  8.177456  8.195863  8.213711  8.242670  8.27636
#> 14 27.439747 27.439747 27.439747 27.477048 27.513215 27.565037 27.62327
#> 15 23.141257 23.141257 23.141257 23.208353 23.273408 23.307291 23.32522
#> 16 17.806073 17.806073 17.806073 17.757740 17.710878 17.720730 17.75794
#> 17 18.235922 18.235922 18.235922 18.184610 18.134858 18.139051 18.16938
#> 18 24.784798 24.784798 24.784798 24.688093 24.594328 24.610021 24.67859
#> 19 27.955566 27.955566 27.955566 27.989292 28.021992 27.965997 27.86817
#> 20 21.598858 21.598858 21.598858 21.524236 21.451883 21.370401 21.28638
#> 21 17.553221 17.553221 17.553221 17.506641 17.461477 17.168269 16.76125
#>       0.3681    0.3433    0.3202    0.2986    0.2785    0.2597   0.2422
#> 1  22.334667 22.388426 22.429261 22.468470 22.505392 22.539499 22.54135
#> 2  21.119539 21.195700 21.275635 21.351039 21.421744 21.487327 21.51615
#> 3  24.058285 24.071774 24.086669 24.100112 24.112562 24.124249 24.14346
#> 4  19.685714 19.705720 19.638903 19.577546 19.520336 19.466984 19.42100
#> 5  18.015362 17.975714 17.878467 17.789045 17.705721 17.627963 17.55028
#> 6  18.549918 18.585193 18.477544 18.379143 18.287507 18.201949 18.11808
#> 7  18.531175 18.587610 18.761222 18.920454 19.068738 19.207196 19.35250
#> 8  18.531175 18.587610 18.761222 18.920454 19.068738 19.207196 19.35250
#> 9  14.981601 14.983451 14.956683 14.932482 14.910024 14.888983 14.86061
#> 10 16.601772 16.573753 16.494850 16.422389 16.354887 16.291879 16.22755
#> 11 16.363511 16.339885 16.268649 16.203285 16.142407 16.085570 16.02653
#> 12  9.200258  9.249599  9.341092  9.425783  9.504832  9.578481  9.64106
#> 13  8.307753  8.356243  8.482270  8.598467  8.706825  8.807868  8.90086
#> 14 27.677530 27.763941 27.995735 28.209357 28.408547 28.594307 28.76690
#> 15 23.341986 23.352813 23.346572 23.340589 23.334933 23.329729 23.33299
#> 16 17.792543 17.789460 17.628663 17.481967 17.345500 17.217950 17.07510
#> 17 18.197586 18.206635 18.134713 18.069574 18.009089 17.952455 17.87969
#> 18 24.742349 24.817919 24.943474 25.060774 25.170528 25.272541 25.33524
#> 19 27.777089 27.676354 27.525731 27.385426 27.254267 27.132249 27.04553
#> 20 21.208050 21.133163 21.055995 20.984298 20.917350 20.854998 20.80176
#> 21 16.382135 16.059038 15.916649 15.779348 15.650074 15.530627 15.51086
#>       0.2259    0.2106    0.1964    0.1832   0.1709    0.1593    0.1486
#> 1  22.490551 21.948103 21.472801 21.438909 21.40971 21.382666 21.357466
#> 2  21.476044 20.975298 20.680286 20.647698 20.61888 20.592125 20.567189
#> 3  24.176655 24.588392 25.035484 25.055637 25.07201 25.087091 25.101136
#> 4  19.433868 19.427670 19.392544 19.434060 19.47226 19.507847 19.541028
#> 5  17.500697 18.125052 17.917184 17.882735 17.85111 17.821650 17.794180
#> 6  18.111395 17.767251 17.989371 18.038445 18.08275 18.123942 18.162347
#> 7  19.453044 18.965153 18.973830 19.030045 19.08251 19.131452 19.177095
#> 8  19.453044 18.965153 18.973830 19.030045 19.08251 19.131452 19.177095
#> 9  14.832992 15.037114 14.891709 14.872631 14.85527 14.839111 14.824045
#> 10 16.185669 16.334188 15.948397 15.927579 15.90971 15.893165 15.877747
#> 11 15.986746 16.143442 15.793002 15.772439 15.75464 15.738157 15.722791
#> 12  9.662905  9.714890 10.036562 10.036848 10.03565 10.034414 10.033254
#> 13  8.940793  9.209854  9.636167  9.635963  9.63377  9.631566  9.629499
#> 14 28.843480 29.224267 28.485358 28.491198 28.49906 28.506595 28.513641
#> 15 23.361319 23.695027 22.966838 22.992303 23.01827 23.042678 23.065451
#> 16 16.983467 16.935334 16.886435 16.860079 16.83556 16.812691 16.791362
#> 17 17.817934 18.144126 17.932723 17.898249 17.86661 17.837151 17.809676
#> 18 25.300651 25.987502 27.158779 27.089233 27.01935 26.953762 26.892566
#> 19 27.047593 26.982497 27.793255 27.819811 27.84162 27.861702 27.880414
#> 20 20.764368 19.734629 19.948162 19.940080 19.93321 19.926848 19.920921
#> 21 15.576785 15.495059 15.487283 15.506011 15.52555 15.543935 15.561096
#>       0.1386   0.1292    0.1205    0.1124    0.1048    0.0978    0.0912
#> 1  21.333721 21.31179 21.291118 21.272054 21.111364 21.111462 21.111550
#> 2  20.543784 20.52209 20.501723 20.482857 20.120451 20.120432 20.120416
#> 3  25.114487 25.12671 25.138352 25.148975 24.990320 24.990125 24.989951
#> 4  19.572006 19.60087 19.627804 19.652908 19.983586 19.983666 19.983737
#> 5  17.768498 17.74461 17.722258 17.701480 18.371419 18.371676 18.371906
#> 6  18.198279 18.23169 18.262937 18.291997 18.032288 18.032163 18.032052
#> 7  19.219688 19.25938 19.296442 19.330964 19.116712 19.116580 19.116463
#> 8  19.219688 19.25938 19.296442 19.330964 19.116712 19.116580 19.116463
#> 9  14.809946 14.79684 14.784574 14.773178 14.244675 14.244564 14.244465
#> 10 15.863196 15.84978 15.837102 15.825441 15.565892 15.565937 15.565977
#> 11 15.708306 15.69494 15.682318 15.670696 15.371596 15.371618 15.371637
#> 12 10.032340 10.03133 10.030567 10.029685 10.397505 10.397500 10.397495
#> 13  9.627804  9.62601  9.624576  9.623007  9.765299  9.765184  9.765082
#> 14 28.520020 28.52614 28.531687 28.537007 29.058145 29.058352 29.058536
#> 15 23.086462 23.10626 23.124510 23.141737 22.999991 23.000070 23.000140
#> 16 16.771427 16.75288 16.735521 16.719393 16.826529 16.826633 16.826726
#> 17 17.783987 17.76009 17.737736 17.716955 17.881981 17.882096 17.882198
#> 18 26.836043 26.78283 26.733752 26.687453 26.460477 26.460111 26.459785
#> 19 27.898136 27.91442 27.929852 27.944013 28.304546 28.304514 28.304485
#> 20 19.915303 19.91015 19.905241 19.900761 19.511916 19.511869 19.511828
#> 21 15.576877 15.59180 15.605488 15.618472 16.168596 16.168867 16.169108
#>       0.0850    0.0793    0.0740    0.0690    0.0643    0.0600    0.0559
#> 1  21.111628 21.111698 21.111760 21.111815 21.427748 21.427907 21.428055
#> 2  20.120401 20.120388 20.120377 20.120367 20.372979 20.373019 20.373056
#> 3  24.989795 24.989657 24.989534 24.989425 24.746298 24.746068 24.745856
#> 4  19.983800 19.983857 19.983907 19.983952 20.125448 20.125578 20.125699
#> 5  18.372111 18.372293 18.372455 18.372600 18.354610 18.354817 18.355009
#> 6  18.031954 18.031865 18.031787 18.031717 18.153390 18.153354 18.153320
#> 7  19.116358 19.116265 19.116182 19.116108 19.201190 19.201105 19.201025
#> 8  19.116358 19.116265 19.116182 19.116108 19.201190 19.201105 19.201025
#> 9  14.244377 14.244299 14.244229 14.244167 13.979527 13.979339 13.979165
#> 10 15.566013 15.566045 15.566073 15.566098 15.385886 15.385857 15.385829
#> 11 15.371655 15.371670 15.371684 15.371697 15.179069 15.179016 15.178967
#> 12 10.397491 10.397487 10.397484 10.397481 10.596555 10.596614 10.596670
#> 13  9.764991  9.764909  9.764837  9.764773  9.759393  9.759292  9.759198
#> 14 29.058700 29.058846 29.058976 29.059092 29.104780 29.104926 29.105061
#> 15 23.000202 23.000258 23.000307 23.000351 22.731497 22.731461 22.731429
#> 16 16.826809 16.826882 16.826948 16.827006 17.008735 17.008890 17.009033
#> 17 17.882289 17.882370 17.882443 17.882507 17.896438 17.896530 17.896616
#> 18 26.459494 26.459235 26.459005 26.458799 26.325268 26.324912 26.324582
#> 19 28.304460 28.304437 28.304417 28.304398 28.338457 28.338490 28.338520
#> 20 19.511791 19.511759 19.511729 19.511703 19.743402 19.743454 19.743502
#> 21 16.169323 16.169514 16.169685 16.169836 15.768138 15.768266 15.768385
#>       0.0522    0.0487    0.0454    0.0423    0.0395    0.0368    0.0343
#> 1  21.428191 21.428318 21.538144 21.537963 21.537795 21.537639 21.537494
#> 2  20.373091 20.373123 20.517518 20.517462 20.517409 20.517359 20.517313
#> 3  24.745658 24.745476 25.006819 25.007042 25.007249 25.007441 25.007617
#> 4  20.125811 20.125915 20.144107 20.143964 20.143833 20.143711 20.143598
#> 5  18.355187 18.355351 18.243560 18.243350 18.243157 18.242977 18.242812
#> 6  18.153289 18.153259 18.240127 18.240155 18.240181 18.240205 18.240228
#> 7  19.200952 19.200883 19.018617 19.018722 19.018821 19.018912 19.018997
#> 8  19.200952 19.200883 19.018617 19.018722 19.018821 19.018912 19.018997
#> 9  13.979003 13.978853 13.989057 13.989259 13.989445 13.989617 13.989777
#> 10 15.385804 15.385780 15.349891 15.349927 15.349960 15.349990 15.350018
#> 11 15.178922 15.178879 15.149769 15.149829 15.149885 15.149936 15.149983
#> 12 10.596721 10.596768 10.667236 10.667166 10.667102 10.667042 10.666987
#> 13  9.759112  9.759031  9.860622  9.860722  9.860815  9.860901  9.860981
#> 14 29.105186 29.105302 29.180140 29.179982 29.179835 29.179698 29.179570
#> 15 22.731399 22.731371 22.675455 22.675500 22.675542 22.675579 22.675613
#> 16 17.009166 17.009288 16.942060 16.941899 16.941750 16.941613 16.941486
#> 17 17.896695 17.896768 17.800660 17.800570 17.800488 17.800411 17.800341
#> 18 26.324276 26.323992 26.195976 26.196370 26.196735 26.197075 26.197390
#> 19 28.338547 28.338573 28.313766 28.313731 28.313700 28.313671 28.313646
#> 20 19.743546 19.743587 19.719119 19.719061 19.719009 19.718960 19.718915
#> 21 15.768495 15.768597 15.828741 15.828601 15.828470 15.828349 15.828236
#>       0.0320    0.0299    0.0278    0.0260    0.0242    0.0226    0.0211
#> 1  21.505472 21.505373 21.505278 21.505189 21.505104 21.505023 21.504946
#> 2  20.535991 20.535980 20.535970 20.535960 20.535950 20.535940 20.535931
#> 3  25.118951 25.119201 25.119437 25.119659 25.119870 25.120069 25.120258
#> 4  20.128035 20.128016 20.127996 20.127978 20.127960 20.127943 20.127926
#> 5  18.269229 18.269207 18.269186 18.269165 18.269146 18.269127 18.269110
#> 6  18.268870 18.268936 18.268998 18.269058 18.269113 18.269166 18.269216
#> 7  19.121289 19.121351 19.121411 19.121468 19.121523 19.121576 19.121626
#> 8  19.121289 19.121351 19.121411 19.121468 19.121523 19.121576 19.121626
#> 9  14.119556 14.119702 14.119840 14.119971 14.120096 14.120214 14.120326
#> 10 15.412197 15.412225 15.412252 15.412277 15.412301 15.412325 15.412347
#> 11 15.222103 15.222148 15.222191 15.222232 15.222271 15.222308 15.222344
#> 12 10.512470 10.512359 10.512254 10.512153 10.512058 10.511968 10.511882
#> 13  9.772039  9.772036  9.772034  9.772031  9.772028  9.772026  9.772023
#> 14 29.281254 29.281307 29.281357 29.281405 29.281449 29.281491 29.281531
#> 15 22.238199 22.237751 22.237327 22.236926 22.236547 22.236188 22.235848
#> 16 16.987373 16.987371 16.987368 16.987366 16.987364 16.987361 16.987359
#> 17 17.929369 17.929450 17.929527 17.929600 17.929669 17.929735 17.929797
#> 18 26.115233 26.115289 26.115342 26.115394 26.115443 26.115490 26.115536
#> 19 28.320480 28.320539 28.320594 28.320646 28.320695 28.320742 28.320785
#> 20 19.574390 19.574203 19.574027 19.573860 19.573702 19.573553 19.573412
#> 21 15.846211 15.846205 15.846199 15.846192 15.846186 15.846179 15.846173
#>       0.0196    0.0183    0.0171    0.0159   0.0149    0.0139    0.0129
#> 1  21.504874 21.504804 21.504739 21.504676 21.50462 21.504561 21.504508
#> 2  20.535922 20.535914 20.535906 20.535898 20.53589 20.535884 20.535877
#> 3  25.120436 25.120605 25.120764 25.120915 25.12106 25.121194 25.121322
#> 4  20.127910 20.127895 20.127881 20.127867 20.12785 20.127842 20.127831
#> 5  18.269093 18.269077 18.269062 18.269048 18.26903 18.269022 18.269010
#> 6  18.269263 18.269307 18.269349 18.269389 18.26943 18.269463 18.269497
#> 7  19.121674 19.121719 19.121763 19.121804 19.12184 19.121881 19.121917
#> 8  19.121674 19.121719 19.121763 19.121804 19.12184 19.121881 19.121917
#> 9  14.120433 14.120534 14.120630 14.120720 14.12081 14.120888 14.120966
#> 10 15.412368 15.412388 15.412407 15.412425 15.41244 15.412458 15.412474
#> 11 15.222377 15.222409 15.222439 15.222468 15.22250 15.222521 15.222546
#> 12 10.511800 10.511723 10.511649 10.511579 10.51151 10.511450 10.511391
#> 13  9.772021  9.772018  9.772015  9.772013  9.77201  9.772008  9.772006
#> 14 29.281568 29.281603 29.281636 29.281667 29.28170 29.281725 29.281751
#> 15 22.235525 22.235220 22.234931 22.234657 22.23440 22.234150 22.233917
#> 16 16.987357 16.987355 16.987353 16.987352 16.98735 16.987349 16.987348
#> 17 17.929856 17.929912 17.929966 17.930016 17.93006 17.930109 17.930153
#> 18 26.115579 26.115620 26.115659 26.115697 26.11573 26.115766 26.115798
#> 19 28.320827 28.320866 28.320903 28.320938 28.32097 28.321003 28.321033
#> 20 19.573278 19.573151 19.573031 19.572917 19.57281 19.572706 19.572610
#> 21 15.846166 15.846160 15.846154 15.846148 15.84614 15.846136 15.846131
#>       0.0121    0.0112    0.0105    0.0098    0.0091    0.0085   0.0079
#> 1  21.504457 21.545676 21.545732 21.545787 21.545839 21.545890 21.54594
#> 2  20.535870 20.495462 20.495436 20.495410 20.495385 20.495361 20.49534
#> 3  25.121443 25.180809 25.180981 25.181148 25.181310 25.181468 25.18162
#> 4  20.127819 20.109240 20.109160 20.109082 20.109007 20.108934 20.10886
#> 5  18.268998 18.272407 18.272328 18.272252 18.272177 18.272105 18.27203
#> 6  18.269529 18.142307 18.142121 18.141941 18.141767 18.141598 18.14143
#> 7  19.121951 19.210219 19.210412 19.210599 19.210781 19.210957 19.21113
#> 8  19.121951 19.068006 19.067978 19.067951 19.067925 19.067900 19.06788
#> 9  14.121039 14.214994 14.215233 14.215465 14.215691 14.215910 14.21612
#> 10 15.412488 15.390900 15.390901 15.390901 15.390902 15.390903 15.39090
#> 11 15.222569 15.116193 15.116071 15.115952 15.115837 15.115726 15.11562
#> 12 10.511334 10.482648 10.482546 10.482447 10.482351 10.482258 10.48217
#> 13  9.772003  9.826903  9.826992  9.827078  9.827161  9.827241  9.82732
#> 14 29.281776 29.255799 29.255775 29.255751 29.255729 29.255707 29.25569
#> 15 22.233695 22.259726 22.259734 22.259742 22.259748 22.259753 22.25976
#> 16 16.987346 17.081023 17.081115 17.081205 17.081291 17.081376 17.08146
#> 17 17.930194 17.925468 17.925447 17.925426 17.925406 17.925387 17.92537
#> 18 26.115829 26.052997 26.052933 26.052872 26.052812 26.052754 26.05270
#> 19 28.321062 28.371664 28.371672 28.371681 28.371689 28.371697 28.37171
#> 20 19.572518 19.583662 19.583644 19.583626 19.583609 19.583591 19.58357
#> 21 15.846126 15.813894 15.813787 15.813683 15.813582 15.813484 15.81339
#>       0.0074    0.0069   0.0064    0.0060    0.0056    0.0052    0.0049
#> 1  21.545986 21.546032 21.54608 21.546119 21.546160 21.546200 21.546239
#> 2  20.495315 20.495293 20.49527 20.495251 20.495231 20.495211 20.495192
#> 3  25.181771 25.181916 25.18206 25.182195 25.182329 25.182459 25.182586
#> 4  20.108794 20.108727 20.10866 20.108599 20.108538 20.108479 20.108421
#> 5  18.271967 18.271901 18.27184 18.271775 18.271715 18.271656 18.271599
#> 6  18.141276 18.141122 18.14097 18.140829 18.140689 18.140553 18.140422
#> 7  19.211294 19.211456 19.21161 19.211764 19.211911 19.212055 19.212193
#> 8  19.067853 19.067830 19.06781 19.067787 19.067767 19.067747 19.067728
#> 9  14.216329 14.216529 14.21672 14.216913 14.217096 14.217275 14.217448
#> 10 15.390904 15.390905 15.39091 15.390907 15.390907 15.390908 15.390909
#> 11 15.115513 15.115412 15.11531 15.115218 15.115126 15.115036 15.114950
#> 12 10.482079 10.481993 10.48191 10.481829 10.481750 10.481674 10.481599
#> 13  9.827395  9.827469  9.82754  9.827609  9.827677  9.827742  9.827805
#> 14 29.255665 29.255646 29.25563 29.255608 29.255591 29.255573 29.255557
#> 15 22.259760 22.259762 22.25976 22.259764 22.259764 22.259763 22.259762
#> 16 17.081536 17.081613 17.08169 17.081760 17.081830 17.081898 17.081964
#> 17 17.925352 17.925335 17.92532 17.925303 17.925288 17.925274 17.925260
#> 18 26.052644 26.052591 26.05254 26.052491 26.052443 26.052397 26.052352
#> 19 28.371713 28.371721 28.37173 28.371736 28.371744 28.371751 28.371758
#> 20 19.583556 19.583539 19.58352 19.583505 19.583489 19.583472 19.583456
#> 21 15.813297 15.813208 15.81312 15.813038 15.812956 15.812877 15.812800
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 22.8 21.4 18.7 18.1 19.2 17.8 16.4 17.3 15.2 10.4 10.4 30.4 21.5
#> [16] 15.5 15.2 26.0 30.4 19.7 15.0
#> 
#> $X
#>               am       carb        cyl        disp        drat       gear
#>  [1,]  1.2747549  0.5819877 -0.2399041 -0.63600888  0.71835779  0.4291975
#>  [2,]  1.2747549  0.5819877 -0.2399041 -0.63600888  0.71835779  0.4291975
#>  [3,]  1.2747549 -1.1639754 -1.4994004 -1.10376326  0.62834876  0.4291975
#>  [4,] -0.7844645 -1.1639754 -0.2399041  0.24552821 -0.75779032 -0.8583951
#>  [5,] -0.7844645 -0.5819877  1.0195922  1.16304642 -0.63177767 -0.8583951
#>  [6,] -0.7844645 -1.1639754 -0.2399041 -0.05131591 -1.33384811 -0.8583951
#>  [7,] -0.7844645  0.5819877 -0.2399041 -0.56764478  0.75436140  0.4291975
#>  [8,] -0.7844645  0.5819877 -0.2399041 -0.56764478  0.75436140  0.4291975
#>  [9,] -0.7844645  0.0000000  1.0195922  0.40564414 -0.77579212 -0.8583951
#> [10,] -0.7844645  0.0000000  1.0195922  0.40564414 -0.77579212 -0.8583951
#> [11,] -0.7844645  0.0000000  1.0195922  0.40564414 -0.77579212 -0.8583951
#> [12,] -0.7844645  0.5819877  1.0195922  2.17051738 -1.02781741 -0.8583951
#> [13,] -0.7844645  0.5819877  1.0195922  2.06257407 -0.90180477 -0.8583951
#> [14,]  1.2747549 -0.5819877 -1.4994004 -1.39431069  2.57254382  0.4291975
#> [15,] -0.7844645 -1.1639754 -1.4994004 -0.99492042  0.35832167 -0.8583951
#> [16,] -0.7844645 -0.5819877  1.0195922  0.78524480 -1.33384811 -0.8583951
#> [17,] -0.7844645 -0.5819877  1.0195922  0.65931093 -0.63177767 -0.8583951
#> [18,]  1.2747549 -0.5819877 -1.4994004 -0.99312136  1.67245351  1.7167902
#> [19,]  1.2747549 -0.5819877 -1.4994004 -1.21980233  0.48433431  1.7167902
#> [20,]  1.2747549  1.7459631 -0.2399041 -0.77093803  0.21430722  1.7167902
#> [21,]  1.2747549  2.9099385  1.0195922  0.63232510  0.07029277  1.7167902
#>                hp        qsec         vs         wt
#>  [1,] -0.61155304 -0.90680663 -0.7844645 -0.6469123
#>  [2,] -0.61155304 -0.48877268 -0.7844645 -0.3819019
#>  [3,] -0.89806805  0.69814513  1.2747549 -0.9586891
#>  [4,] -0.61155304  1.31773115  1.2747549 -0.0285548
#>  [5,]  0.48394552 -0.48877268 -0.7844645  0.2052778
#>  [6,] -0.69582216  1.89999272  1.2747549  0.2260630
#>  [7,] -0.39245333  0.46673348  1.2747549  0.2052778
#>  [8,] -0.39245333  0.91462699  1.2747549  0.2052778
#>  [9,]  0.56821464 -0.20510679 -0.7844645  0.8600093
#> [10,]  0.56821464 -0.05580895 -0.7844645  0.5066622
#> [11,]  0.56821464  0.24278672 -0.7844645  0.5586250
#> [12,]  0.98956024  0.22785694 -0.7844645  2.0863316
#> [13,]  1.15809848  0.10841867 -0.7844645  2.2671622
#> [14,] -1.58907484  0.63096110  1.2747549 -1.6913648
#> [15,] -0.83065276  1.74322999  1.2747549 -0.8079970
#> [16,]  0.06259992 -0.60074606 -0.7844645  0.2884183
#> [17,]  0.06259992 -0.27975571 -0.7844645  0.2000816
#> [18,] -0.93177570 -0.72764922 -0.7844645 -1.1457552
#> [19,] -0.56099157 -0.57835138  1.2747549 -1.7973689
#> [20,]  0.48394552 -1.62343625 -0.7844645 -0.4910238
#> [21,]  3.18055736 -2.29527652 -0.7844645  0.3403812
#> attr(,"center")
#>  [1]   0.3809524   3.0000000   6.3809524 230.7047619   3.5009524   3.6666667
#>  [7] 146.2857143  17.6747619   0.3809524   3.2424762
#> attr(,"scale")
#>  [1]   0.4856209   1.7182494   1.5879364 111.1694566   0.5554998   0.7766432
#>  [7]  59.3337155   1.3396041   0.4856209   0.9622266
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 19
#> 
#> $lambda.min
#> [1] 1.385861
#> 
#> $null.dev
#> [1] 26.66753
#> 
#> $Bias
#> [1] 1.724277
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#>  21.4698