Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 38.419771 36.489565 34.247328 30.342534 26.752133 23.817080 21.439367
#>   [8] 19.534602 18.030181 16.863696 15.981564 15.337821 14.898478 14.524017
#>  [15] 14.255881 14.045245 13.800476 13.374661 12.992631 12.590954 12.201498
#>  [22] 11.902016 11.211224 10.579502 10.057854  9.629498  9.280829  9.196589
#>  [29]  9.696749 10.030268  9.961555 10.250352  9.752739  8.959489  7.999081
#>  [36]  7.763275  7.094281  7.415672  6.804542  6.470845  7.455978  7.512483
#>  [43]  7.512569  8.194234  8.136230  8.712889  9.162762  9.163343  9.212269
#>  [50]  9.187495 11.330507 11.331621 11.015952 11.012104 11.113408 11.102557
#>  [57] 11.103286 11.358327 11.359579 11.280498 11.657198 11.634085 11.909154
#>  [64] 12.113072 12.097185 12.187450 12.521717 12.772535 12.920348 13.258292
#>  [71] 13.380349 13.412202 13.380257 13.381088 13.441048 13.442138 13.443136
#>  [78] 13.448284 13.493817 13.502431 13.521777 14.045919 14.047218 14.012720
#>  [85] 14.013790 14.014786 14.015949 14.142398 14.142951 14.174139 14.178342
#>  [92] 14.178884 14.179416 14.180911 14.180620 14.182522 14.182919 14.183286
#>  [99] 14.183626 14.183941
#> 
#> $cvse
#>   [1] 13.051430 12.948059 12.792789 11.653826 10.562104  9.629023  8.819593
#>   [8]  8.106904  7.470969  6.897529  6.376924  5.903105  5.468342  5.081644
#>  [15]  4.743840  4.454389  4.204868  3.985558  3.829918  3.757994  3.704079
#>  [22]  3.674913  3.371499  3.084814  2.853227  2.671833  2.536018  2.529932
#>  [29]  2.638009  2.746894  2.815390  2.750283  2.690298  2.611637  2.243194
#>  [36]  2.175847  1.806773  1.972493  1.955102  1.842617  1.904975  1.897958
#>  [43]  1.898003  2.448781  2.477342  2.549868  2.533600  2.533665  2.522674
#>  [50]  2.521419  4.427294  4.428134  4.414354  4.415519  4.562017  4.562376
#>  [57]  4.562898  5.032494  5.033305  5.035222  5.023249  5.025610  5.032262
#>  [64]  5.025711  5.035942  5.033045  5.127813  5.109548  5.115237  5.092415
#>  [71]  5.121390  5.120437  5.123046  5.123207  5.125660  5.125781  5.125888
#>  [78]  5.127261  5.126609  5.125531  5.123898  5.550928  5.551682  5.552186
#>  [85]  5.552884  5.553553  5.554251  5.575329  5.575852  5.575524  5.575840
#>  [92]  5.576292  5.576720  5.576895  5.577466  5.577967  5.578342  5.578697
#>  [99]  5.579035  5.579356
#> 
#> $fold
#>  [1]  3  5 10  6  5  9  1  2  4  1  2  1 10  6  8  9  4  3  7  8  7
#> 
#> $lambda
#>   [1] 5.143090684 4.796463585 4.473197993 4.171719419 3.890559493 3.628348804
#>   [7] 3.383810238 3.155752754 2.943065580 2.744712810 2.559728353 2.387211229
#>  [13] 2.226321182 2.076274585 1.936340627 1.805837750 1.684130329 1.570625582
#>  [19] 1.464770674 1.366050033 1.273982833 1.188120654 1.108045298 1.033366754
#>  [25] 0.963721293 0.898769703 0.838195633 0.781704053 0.729019816 0.679886321
#>  [31] 0.634064258 0.591330449 0.551476756 0.514309068 0.479646359 0.447319799
#>  [37] 0.417171942 0.389055949 0.362834880 0.338381024 0.315575275 0.294306558
#>  [43] 0.274471281 0.255972835 0.238721123 0.222632118 0.207627459 0.193634063
#>  [49] 0.180583775 0.168413033 0.157062558 0.146477067 0.136605003 0.127398283
#>  [55] 0.118812065 0.110804530 0.103336675 0.096372129 0.089876970 0.083819563
#>  [61] 0.078170405 0.072901982 0.067988632 0.063406426 0.059133045 0.055147676
#>  [67] 0.051430907 0.047964636 0.044731980 0.041717194 0.038905595 0.036283488
#>  [73] 0.033838102 0.031557528 0.029430656 0.027447128 0.025597284 0.023872112
#>  [79] 0.022263212 0.020762746 0.019363406 0.018058377 0.016841303 0.015706256
#>  [85] 0.014647707 0.013660500 0.012739828 0.011881207 0.011080453 0.010333668
#>  [91] 0.009637213 0.008987697 0.008381956 0.007817041 0.007290198 0.006798863
#>  [97] 0.006340643 0.005913304 0.005514768 0.005143091
#> 
#> $fit
#> $beta
#>               5.1431     4.7965    4.4732   4.1717    3.8906    3.6283
#> (Intercept) 19.52381 21.5436158 23.427294 25.18402 26.822346 28.350255
#> am           0.00000  0.0000000  0.000000  0.00000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.000000  0.00000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.000000  0.00000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.000000  0.00000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.000000  0.00000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.000000  0.00000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.000000  0.00000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.000000  0.00000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.000000  0.00000  0.000000  0.000000
#> wt           0.00000 -0.6005456 -1.160616 -1.68294 -2.170061 -2.624352
#>                3.3838    3.1558    2.9431    2.7447   2.5597    2.3872
#> (Intercept) 29.775188 31.104086 32.343420 33.499228 34.57714 35.582400
#> am           0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> wt          -3.048025 -3.443144 -3.811633 -4.155287 -4.47578 -4.774673
#>                2.2263    2.0763    1.9363      1.8058     1.6841     1.5706
#> (Intercept) 36.519911 37.394237 38.209636 37.52959673 36.6946993 34.9337486
#> am           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> carb         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> cyl          0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> disp         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> qsec         0.000000  0.000000  0.000000  0.07148946  0.1448967  0.2357511
#> vs           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> wt          -5.053422 -5.313384 -5.555825 -5.73298597 -5.8742799 -5.8328149
#>                 1.4648     1.3661    1.2740     1.1881     1.1080     1.0334
#> (Intercept) 33.2910490 31.7590617 30.330325 28.9978808 27.7552387 26.5963465
#> am           0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> qsec         0.3205045  0.3995458  0.473260  0.5420061  0.6061189  0.6659108
#> vs           0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> wt          -5.7941343 -5.7580607 -5.724418 -5.6930433 -5.6637829 -5.6364945
#>                 0.9637     0.8988     0.8382     0.7817    0.7290     0.6799
#> (Intercept) 25.5155597 24.5076144 23.5676012 22.6909417 21.873366 21.1108925
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> qsec         0.7216728  0.7736767  0.8221757  0.8674061  0.909588  0.9489271
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> wt          -5.6110453 -5.5873112 -5.5651768 -5.5445341 -5.525283 -5.5073288
#>                0.6341     0.5913    0.5515     0.5143     0.4796     0.4473
#> (Intercept) 15.728951 10.4399689 11.404731 11.6491827 11.6466031 11.6448261
#> am           1.316276  4.6032687  5.004181  5.1047130  5.1055612  5.1061424
#> carb         0.000000 -0.6492687 -0.880274 -0.9384264 -0.9385066 -0.9385608
#> cyl          0.000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> disp         0.000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> drat         0.000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> gear         0.000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> hp           0.000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> qsec         1.131661  1.1870850  1.115377  1.0972716  1.0973459  1.0973973
#> vs           0.000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> wt          -5.007251 -3.4662337 -3.202978 -3.1368554 -3.1364943 -3.1362472
#>                 0.4172     0.3891     0.3628     0.3384     0.3156     0.2943
#> (Intercept) 11.6436056 11.6427690 11.6421965 11.6418051 11.6415376 11.6413550
#> am           5.1065401  5.1068120  5.1069978  5.1071246  5.1072112  5.1072702
#> carb        -0.9385974 -0.9386223 -0.9386392 -0.9386507 -0.9386585 -0.9386638
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         1.0974327  1.0974570  1.0974737  1.0974851  1.0974929  1.0974982
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -3.1360783 -3.1359629 -3.1358842 -3.1358304 -3.1357937 -3.1357687
#>                 0.2745     0.2560     0.2387     0.2226      0.2076      0.1936
#> (Intercept) 11.6412304 11.6411453 16.4699746 16.4725621 21.76773926 21.76730274
#> am           5.1073105  5.1073381  5.6796003  5.6800399  5.60553800  5.60477474
#> carb        -0.9386674 -0.9386699 -1.1258791 -1.1260524 -0.93433300 -0.93406483
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000 -0.01538017 -0.01538343
#> qsec         1.0975018  1.0975043  0.6723334  0.6721213  0.39601498  0.39612979
#> vs           0.0000000  0.0000000  2.2136137  2.2144360  2.23909299  2.23837512
#> wt          -3.1357516 -3.1357400 -2.4531445 -2.4527679 -2.02540103 -2.02581670
#>                  0.1806      0.1684      0.1571      0.1465      0.1366
#> (Intercept) 21.76649110 21.76546419 21.76433405 21.76317759 21.76204607
#> am           5.60415624  5.60365544  5.60325047  5.60292352  5.60266009
#> carb        -0.93385462 -0.93369064 -0.93356336 -0.93346507 -0.93338958
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.01538521 -0.01538594 -0.01538594 -0.01538547 -0.01538469
#> qsec         0.39624971  0.39636855  0.39648227  0.39658845  0.39668578
#> vs           2.23774610  2.23719859  2.23672478  2.23631683  2.23596721
#> wt          -2.02618809 -2.02651693 -2.02680579 -2.02705770 -2.02727599
#>                 0.1274      0.1188      0.1108     0.1033      0.0964
#> (Intercept) 25.1511584 25.15250725 25.15372500 25.1548721 25.15597508
#> am           5.9516988  5.95227723  5.95274201  5.9531172  5.95342086
#> carb        -0.8933910 -0.89363339 -0.89382696 -0.8939785 -0.89409515
#> cyl          0.0000000  0.00000000  0.00000000  0.0000000  0.00000000
#> disp         0.0000000  0.00000000  0.00000000  0.0000000  0.00000000
#> drat        -0.8286715 -0.82864666 -0.82858038 -0.8285013 -0.82842533
#> gear         0.0000000  0.00000000  0.00000000  0.0000000  0.00000000
#> hp          -0.0157167 -0.01571441 -0.01571339 -0.0157133 -0.01571382
#> qsec         0.3860346  0.38588110  0.38573758  0.3856052  0.38548463
#> vs           2.5478732  2.54855669  2.54912818  2.5496107  2.55002071
#> wt          -2.1890588 -2.18869219 -2.18834487 -2.1880266 -2.18774220
#>                  0.0899      0.0838      0.0782      0.0729      0.0680
#> (Intercept) 25.15704038 25.15806358 25.15903588 25.15994804 25.16079250
#> am           5.95366667  5.95386560  5.95402647  5.95415641  5.95426129
#> carb        -0.89418364 -0.89424987 -0.89429878 -0.89433441 -0.89435998
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat        -0.82836001 -0.82830812 -0.82826961 -0.82824301 -0.82822629
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.01571473 -0.01571586 -0.01571709 -0.01571833 -0.01571954
#> qsec         0.38537578  0.38527844  0.38519209  0.38511608  0.38504962
#> vs           2.55037057  2.55066955  2.55092502  2.55114301  2.55132863
#> wt          -2.18749280 -2.18727732 -2.18709329 -2.18693756 -2.18680672
#>                  0.0634      0.0591      0.0551      0.0514      0.0480
#> (Intercept) 25.16156438 25.16226160 25.16288465 25.16343605 25.16391986
#> am           5.95434587  5.95441403  5.95446894  5.95451316  5.95454876
#> carb        -0.89437801 -0.89439046 -0.89439881 -0.89440419 -0.89440745
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat        -0.82821732 -0.82821415 -0.82821513 -0.82821890 -0.82822442
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.01572067 -0.01572171 -0.01572264 -0.01572348 -0.01572421
#> qsec         0.38499190  0.38494205  0.38489924  0.38486266  0.38483153
#> vs           2.55148627  2.55161975  2.55173245  2.55182733  2.55190699
#> wt          -2.18669741 -2.18660652 -2.18653122 -2.18646902 -2.18641777
#>                  0.0447     0.0417      0.0389      0.0363      0.0338
#> (Intercept) 25.16434114 25.1647055 25.16501869 25.16528650 25.16551440
#> am           5.95457742  5.9546005  5.95461905  5.95463398  5.95464599
#> carb        -0.89440921 -0.8944100 -0.89441002 -0.89440964 -0.89440899
#> cyl          0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> drat        -0.82823091 -0.8282378 -0.82824465 -0.82825125 -0.82825742
#> gear         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> hp          -0.01572485 -0.0157254 -0.01572588 -0.01572629 -0.01572663
#> qsec         0.38480516  0.3847829  0.38476417  0.38474847  0.38473534
#> vs           2.55197368  2.5520294  2.55207579  2.55211438  2.55214640
#> wt          -2.18637563 -2.1863410 -2.18631271 -2.18628952 -2.18627057
#>                  0.0316      0.0294      0.0274      0.0256     0.0239
#> (Intercept) 25.16570752 25.16587053 25.16600766 25.16612265 25.1662188
#> am           5.95465565  5.95466341  5.95466965  5.95467466  5.9546787
#> carb        -0.89440820 -0.89440736 -0.89440651 -0.89440569 -0.8944049
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> drat        -0.82826306 -0.82826814 -0.82827266 -0.82827663 -0.8282801
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> hp          -0.01572693 -0.01572717 -0.01572738 -0.01572756 -0.0157277
#> qsec         0.38472439  0.38471528  0.38470772  0.38470145  0.3846963
#> vs           2.55217291  2.55219483  2.55221292  2.55222783  2.5522401
#> wt          -2.18625511 -2.18624250 -2.18623223 -2.18622388 -2.1862171
#>                  0.0223      0.0208      0.0194      0.0181      0.0168
#> (Intercept) 25.16629899 25.16636571 25.16642110 25.16646699 25.16650494
#> am           5.95468192  5.95468452  5.95468660  5.95468827  5.95468960
#> carb        -0.89440424 -0.89440362 -0.89440307 -0.89440259 -0.89440217
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat        -0.82828309 -0.82828566 -0.82828787 -0.82828974 -0.82829132
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.01572782 -0.01572792 -0.01572801 -0.01572808 -0.01572813
#> qsec         0.38469199  0.38468847  0.38468556  0.38468318  0.38468123
#> vs           2.55225017  2.55225844  2.55226522  2.55227077  2.55227530
#> wt          -2.18621157 -2.18620709 -2.18620346 -2.18620052 -2.18619814
#>                  0.0157      0.0146      0.0137      0.0127      0.0119
#> (Intercept) 25.16653627 25.16656208 25.29923802 25.29718542 25.29540162
#> am           5.95469067  5.95469153  6.02047602  6.02042864  6.02036858
#> carb        -0.89440181 -0.89440150 -0.86867813 -0.86844131 -0.86824581
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat        -0.82829266 -0.82829378 -0.76104486 -0.76049241 -0.76005461
#> gear         0.00000000  0.00000000 -0.11458240 -0.11504896 -0.11542332
#> hp          -0.01572818 -0.01572822 -0.01584936 -0.01585051 -0.01585124
#> qsec         0.38467962  0.38467831  0.38570265  0.38581742  0.38592167
#> vs           2.55227901  2.55228203  2.55780368  2.55737115  2.55698821
#> wt          -2.18619621 -2.18619466 -2.20087294 -2.20105141 -2.20122047
#>                  0.0111        0.0103      0.0096      0.0090      0.0084
#> (Intercept) 25.29384201 25.2910851694 24.17039624 24.16595490 24.16226224
#> am           6.02030344  6.0202380832  6.03771150  6.03785870  6.03798697
#> carb        -0.86808480 -0.8679525033 -0.85981234 -0.85989646 -0.85997736
#> cyl          0.00000000  0.0001034969  0.08154022  0.08183041  0.08206873
#> disp         0.00000000  0.0000000000  0.00000000  0.00000000  0.00000000
#> drat        -0.75970674 -0.7591865291 -0.70247013 -0.70272977 -0.70297463
#> gear        -0.11572435 -0.1159972964 -0.08991359 -0.08960765 -0.08933768
#> hp          -0.01585163 -0.0158532521 -0.01664699 -0.01664469 -0.01664239
#> qsec         0.38601602  0.3861113778  0.41408964  0.41426433  0.41441238
#> vs           2.55665062  2.5563504026  2.56684458  2.56697116  2.56708429
#> wt          -2.20137784 -2.2015164511 -2.23420456 -2.23448183 -2.23471743
#>                  0.0078      0.0073      0.0068        0.0063       0.0059
#> (Intercept) 24.15922864 24.15677211 24.15481801  2.415467e+01 24.266532853
#> am           6.03809813  6.03819393  6.03827597  6.038171e+00  6.059901374
#> carb        -0.86005454 -0.86012764 -0.86019640 -8.602576e-01 -0.915178329
#> cyl          0.08226175  0.08241539  0.08253504  8.278281e-02  0.090490528
#> disp         0.00000000  0.00000000  0.00000000 -5.611537e-06 -0.001331293
#> drat        -0.70320421 -0.70341832 -0.70361698 -7.045848e-01 -0.734592318
#> gear        -0.08910052 -0.08889318 -0.08871286 -8.842755e-02 -0.061864444
#> hp          -0.01664013 -0.01663792 -0.01663580 -1.662919e-02 -0.015501255
#> qsec         0.41453673  0.41464010  0.41472498  4.148424e-01  0.398296163
#> vs           2.56718514  2.56727479  2.56735425  2.567417e+00  2.565952380
#> wt          -2.23491592 -2.23508152 -2.23521813 -2.235390e+00 -2.100443405
#>                   0.0055       0.0051
#> (Intercept) 24.266219305 24.265918929
#> am           6.059994902  6.060085087
#> carb        -0.915401208 -0.915616132
#> cyl          0.090577300  0.090660845
#> disp        -0.001335895 -0.001340332
#> drat        -0.734639212 -0.734684583
#> gear        -0.061682409 -0.061506926
#> hp          -0.015498453 -0.015495750
#> qsec         0.398225907  0.398158123
#> vs           2.566031739  2.566108278
#> wt          -2.099903232 -2.099382364
#> 
#> $iter
#>   [1]  0  4  3  3  3  2  2  2  2  2  2  2  2  2  2  6  4  3  3  3  3  3  3  3  3
#>  [26]  3  3  3  3  3 43 95 33 12  1  1  1  1  1  1  1  1  1  1 41  1 39  1  1  1
#>  [51]  1  1  1 42  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#>  [76]  1  1  1  1  1  1  1  1  1  1 26  1  1  1  1 34  1  1  1  1  1  2 72  1  1
#> 
#> $lambda
#>   [1] 5.143090684 4.796463585 4.473197993 4.171719419 3.890559493 3.628348804
#>   [7] 3.383810238 3.155752754 2.943065580 2.744712810 2.559728353 2.387211229
#>  [13] 2.226321182 2.076274585 1.936340627 1.805837750 1.684130329 1.570625582
#>  [19] 1.464770674 1.366050033 1.273982833 1.188120654 1.108045298 1.033366754
#>  [25] 0.963721293 0.898769703 0.838195633 0.781704053 0.729019816 0.679886321
#>  [31] 0.634064258 0.591330449 0.551476756 0.514309068 0.479646359 0.447319799
#>  [37] 0.417171942 0.389055949 0.362834880 0.338381024 0.315575275 0.294306558
#>  [43] 0.274471281 0.255972835 0.238721123 0.222632118 0.207627459 0.193634063
#>  [49] 0.180583775 0.168413033 0.157062558 0.146477067 0.136605003 0.127398283
#>  [55] 0.118812065 0.110804530 0.103336675 0.096372129 0.089876970 0.083819563
#>  [61] 0.078170405 0.072901982 0.067988632 0.063406426 0.059133045 0.055147676
#>  [67] 0.051430907 0.047964636 0.044731980 0.041717194 0.038905595 0.036283488
#>  [73] 0.033838102 0.031557528 0.029430656 0.027447128 0.025597284 0.023872112
#>  [79] 0.022263212 0.020762746 0.019363406 0.018058377 0.016841303 0.015706256
#>  [85] 0.014647707 0.013660500 0.012739828 0.011881207 0.011080453 0.010333668
#>  [91] 0.009637213 0.008987697 0.008381956 0.007817041 0.007290198 0.006798863
#>  [97] 0.006340643 0.005913304 0.005514768 0.005143091
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 30
#> 
#> $loss
#>   [1] 746.61810 639.98292 550.76674 476.46282 414.90699 364.23187 322.82729
#>   [8] 289.30581 262.47281 241.30060 224.90589 212.53029 203.52333 197.32786
#>  [15] 193.46727 180.87840 170.21833 158.81343 149.36470 141.58821 135.23637
#>  [22] 130.09588 125.98308 122.73998 120.23077 118.33889 116.96434 116.02148
#>  [29] 115.43706 115.14850 100.55814  69.66631  67.47600  67.38302  67.38301
#>  [36]  67.38301  67.38301  67.38301  67.38301  67.38301  67.38301  67.38301
#>  [43]  67.38301  67.38301  61.66471  61.66470  56.94248  56.94246  56.94245
#>  [50]  56.94245  56.94244  56.94244  56.94243  55.80666  55.80664  55.80663
#>  [57]  55.80663  55.80662  55.80662  55.80662  55.80662  55.80661  55.80661
#>  [64]  55.80661  55.80661  55.80661  55.80661  55.80661  55.80661  55.80661
#>  [71]  55.80661  55.80661  55.80661  55.80661  55.80661  55.80661  55.80661
#>  [78]  55.80661  55.80661  55.80661  55.80661  55.80661  55.80661  55.80661
#>  [85]  55.80661  55.78015  55.78014  55.78013  55.78012  55.78004  55.74902
#>  [92]  55.74901  55.74900  55.74900  55.74899  55.74899  55.74880  55.72442
#>  [99]  55.72440  55.72438
#> 
#> $linear.predictors
#>      5.1431   4.7965   4.4732   4.1717   3.8906   3.6283   3.3838   3.1558
#> 1  19.52381 19.81705 20.09052 20.34557 20.58342 20.80524 21.01212 21.20505
#> 2  19.52381 19.61286 19.69591 19.77337 19.84560 19.91296 19.97579 20.03438
#> 3  19.52381 19.47774 19.43477 19.39470 19.35733 19.32248 19.28998 19.25967
#> 4  19.52381 19.62788 19.72493 19.81544 19.89985 19.97857 20.05199 20.12046
#> 5  19.52381 19.65190 19.77135 19.88276 19.98665 20.08355 20.17391 20.25818
#> 6  19.52381 19.47774 19.43477 19.39470 19.35733 19.32248 19.28998 19.25967
#> 7  19.52381 19.47774 19.43477 19.39470 19.35733 19.32248 19.28998 19.25967
#> 8  19.52381 19.09940 18.70359 18.33445 17.99020 17.66914 17.36973 17.09049
#> 9  19.52381 19.27355 19.04016 18.82250 18.61951 18.43020 18.25365 18.08900
#> 10 19.52381 18.39075 17.33406 16.34858 15.42952 14.57241 13.77306 13.02758
#> 11 19.52381 18.28626 17.13211 16.05575 15.05193 14.11577 13.24270 12.42847
#> 12 19.52381 20.22242 20.87394 21.48155 22.04821 22.57668 23.06953 23.52917
#> 13 19.52381 20.44161 21.29756 22.09582 22.84028 23.53457 24.18206 24.78592
#> 14 19.52381 20.06327 20.56637 21.03557 21.47314 21.88123 22.26181 22.61674
#> 15 19.52381 19.42970 19.34192 19.26007 19.18373 19.11254 19.04614 18.98422
#> 16 19.52381 19.23752 18.97053 18.72153 18.48931 18.27274 18.07077 17.88241
#> 17 19.52381 19.23452 18.96472 18.71311 18.47846 18.25962 18.05553 17.86520
#> 18 19.52381 20.25845 20.94357 21.58253 22.17841 22.73414 23.25241 23.73576
#> 19 19.52381 19.63989 19.74814 19.84910 19.94325 20.03106 20.11295 20.18932
#> 20 19.52381 19.88010 20.21239 20.52227 20.81128 21.08080 21.33216 21.56658
#> 21 19.52381 19.39967 19.28389 19.17592 19.07523 18.98132 18.89374 18.81206
#>      2.9431   2.7447   2.5597    2.3872    2.2263    2.0763    1.9363    1.8058
#> 1  21.38498 21.55278 21.70927 21.855214 21.991323 22.118259 22.236640 22.264013
#> 2  20.08902 20.13998 20.18750 20.231825 20.273160 20.311708 20.347659 20.487802
#> 3  19.23140 19.20504 19.18045 19.157524 19.136140 19.116197 19.097598 19.024876
#> 4  20.18431 20.24386 20.29940 20.351192 20.399495 20.444543 20.486555 20.671161
#> 5  20.33678 20.41007 20.47843 20.542179 20.601632 20.657078 20.708788 21.107800
#> 6  19.23140 19.20504 19.18045 19.157524 19.136140 19.116197 19.097598 19.116382
#> 7  19.23140 19.20504 19.18045 19.157524 19.136140 19.116197 19.097598 19.159276
#> 8  16.83007 16.58721 16.36071 16.149480 15.952484 15.768765 15.597429 15.440260
#> 9  17.93545 17.79224 17.65869 17.534135 17.417976 17.309647 17.208618 17.145720
#> 10 12.33235 11.68397 11.07929 10.515365  9.989446  9.498973  9.041555  8.716801
#> 11 11.66912 10.96095 10.30050  9.684572  9.110151  8.574444  8.074842  7.707823
#> 12 23.95783 24.35760 24.73042 25.078119 25.402383 25.704793 25.986821 26.308927
#> 13 25.34907 25.87428 26.36408 26.820875 27.246882 27.644178 28.014697 28.432208
#> 14 22.94775 23.25644 23.54434 23.812830 24.063226 24.296746 24.514528 24.828290
#> 15 18.92647 18.87262 18.82239 18.775550 18.731866 18.691126 18.653132 18.555513
#> 16 17.70675 17.54292 17.39014 17.247655 17.114771 16.990844 16.875268 16.616583
#> 17 17.68769 17.52215 17.36776 17.223781 17.089504 16.964277 16.847489 16.705161
#> 18 24.18653 24.60691 24.99897 25.364599 25.705588 26.023596 26.320171 26.454881
#> 19 20.26054 20.32697 20.38891 20.446686 20.500564 20.550811 20.597671 20.392628
#> 20 21.78520 21.98908 22.17923 22.356555 22.521933 22.676164 22.820001 22.757312
#> 21 18.73589 18.66485 18.59860 18.536816 18.479195 18.425457 18.375341 18.106583
#>       1.6841    1.5706    1.4648    1.3661    1.2740    1.1881    1.1080
#> 1  22.272286 22.176890 22.087900 22.004907 21.927508 21.855325 21.788007
#> 2  20.625681 20.764251 20.893515 21.014067 21.126495 21.231345 21.329128
#> 3  18.953318 18.881349 18.814214 18.751603 18.693211 18.638755 18.587970
#> 4  20.853680 21.042092 21.217851 21.381764 21.534631 21.677194 21.810150
#> 5  21.508852 21.959082 22.379079 22.770770 23.136061 23.476734 23.794446
#> 6  19.138786 19.183111 19.224459 19.263021 19.298984 19.332523 19.363802
#> 7  19.225724 19.324562 19.416762 19.502749 19.582940 19.657727 19.727473
#> 8  15.307582 15.296261 15.285701 15.275852 15.266666 15.258100 15.250112
#> 9  17.098062 17.129228 17.158302 17.185417 17.210704 17.234287 17.256280
#> 10  8.459972  8.550276  8.634515  8.713077  8.786344  8.854673  8.918397
#> 11  7.414664  7.497646  7.575055  7.647247  7.714573  7.777362  7.835919
#> 12 26.592422 26.691630 26.784176 26.870485 26.950977 27.026044 27.096052
#> 13 28.798840 28.921981 29.036852 29.143982 29.243891 29.337067 29.423964
#> 14 25.113982 25.273240 25.421803 25.560354 25.689566 25.810071 25.922453
#> 15 18.461641 18.379362 18.302607 18.231026 18.164269 18.102011 18.043949
#> 16 16.370322 16.168664 15.980548 15.805109 15.641495 15.488908 15.346605
#> 17 16.578582 16.526132 16.477204 16.431574 16.389020 16.349333 16.312321
#> 18 26.543515 26.388568 26.244027 26.109227 25.983512 25.866270 25.756929
#> 19 20.174234 19.862117 19.570959 19.299423 19.046189 18.810022 18.589771
#> 20 22.668843 22.430994 22.209117 22.002194 21.809216 21.629245 21.461403
#> 21 17.839012 17.552566 17.285355 17.036154 16.803748 16.587005 16.384870
#>       1.0334   0.9637    0.8988    0.8382    0.7817    0.7290    0.6799
#> 1  21.725226 21.66668 21.612073 21.561149 21.513658 21.469367 21.428061
#> 2  21.420322 21.50537 21.584684 21.658654 21.727639 21.791974 21.851973
#> 3  18.540607 18.49644 18.455242 18.416824 18.380996 18.347582 18.316420
#> 4  21.934144 22.04978 22.157626 22.258202 22.351999 22.439475 22.521055
#> 5  24.090745 24.36708 24.624781 24.865119 25.089258 25.298292 25.493237
#> 6  19.392972 19.42018 19.445548 19.469209 19.491276 19.511855 19.531047
#> 7  19.792519 19.85318 19.909754 19.962515 20.011719 20.057608 20.100403
#> 8  15.242661 15.23571 15.229233 15.223190 15.217554 15.212298 15.207396
#> 9  17.276791 17.29592 17.313759 17.330396 17.345912 17.360382 17.373877
#> 10  8.977826  9.03325  9.084938  9.133143  9.178099  9.220025  9.259125
#> 11  7.890530  7.94146  7.988958  8.033254  8.074565  8.113092  8.149022
#> 12 27.161341 27.22223 27.279016 27.331974 27.381363 27.427423 27.470380
#> 13 29.505003 29.58058 29.651065 29.716799 29.778103 29.835274 29.888593
#> 14 26.027262 26.12501 26.216164 26.301177 26.380461 26.454401 26.523358
#> 15 17.989800 17.93930 17.892205 17.848284 17.807322 17.769121 17.733495
#> 16 15.213892 15.09012 14.974698 14.867051 14.766658 14.673032 14.585716
#> 17 16.277804 16.24561 16.215591 16.187593 16.161482 16.137130 16.114420
#> 18 25.654958 25.55986 25.471170 25.388458 25.311320 25.239382 25.172291
#> 19 18.384365 18.19280 18.014150 17.847539 17.692157 17.547247 17.412103
#> 20 21.304874 21.15889 21.022752 20.895785 20.777376 20.666948 20.563962
#> 21 16.196358 16.02055 15.856594 15.703686 15.561084 15.428092 15.304064
#>       0.6341   0.5913   0.5515   0.5143   0.4796   0.4473   0.4172   0.3891
#> 1  21.910250 22.68493 22.66297 22.65729 22.65754 22.65772 22.65784 22.65792
#> 2  21.630129 21.72369 21.90981 21.95673 21.95667 21.95663 21.95661 21.95659
#> 3  17.764878 17.42177 17.60965 17.65711 17.65688 17.65672 17.65661 17.65653
#> 4  22.389040 21.82585 21.73422 21.71119 21.71109 21.71102 21.71098 21.71094
#> 5  25.871147 25.40704 25.09693 25.01875 25.01885 25.01892 25.01897 25.01901
#> 6  19.213404 17.64271 17.27679 17.18476 17.18447 17.18426 17.18413 17.18403
#> 7  19.892401 18.35496 17.94601 17.84313 17.84287 17.84270 17.84259 17.84251
#> 8  15.040341 15.03987 15.13534 15.15943 15.15937 15.15933 15.15930 15.15929
#> 9  17.171441 16.75733 16.73343 16.72748 16.72736 16.72728 16.72723 16.72719
#> 10  9.788149 10.98896 11.12247 11.15593 11.15626 11.15649 11.15665 11.15675
#> 11  8.735821 10.19590 10.38670 10.43455 10.43494 10.43520 10.43538 10.43550
#> 12 28.062714 29.88080 30.19847 30.27827 30.27870 30.27899 30.27919 30.27933
#> 13 30.376975 31.65642 31.84717 31.89504 31.89537 31.89560 31.89576 31.89586
#> 14 26.030614 25.00000 24.94780 24.93481 24.93453 24.93434 24.93420 24.93411
#> 15 17.194549 16.96641 17.18611 17.24157 17.24136 17.24121 17.24111 17.24103
#> 16 13.940003 12.82554 12.77216 12.75891 12.75854 12.75829 12.75811 12.75799
#> 17 15.770891 16.05356 16.34591 16.41960 16.41952 16.41946 16.41942 16.41939
#> 18 25.228448 26.15128 26.42078 26.48861 26.48873 26.48881 26.48887 26.48891
#> 19 17.581326 18.67093 18.90734 18.96680 18.96697 18.96708 18.96716 18.96722
#> 20 20.715887 19.94598 19.54336 19.44196 19.44190 19.44186 19.44183 19.44181
#> 21 15.691591 14.80608 14.21659 14.06808 14.06808 14.06808 14.06808 14.06809
#>      0.3628   0.3384   0.3156   0.2943   0.2745   0.2560   0.2387   0.2226
#> 1  22.65797 22.65801 22.65804 22.65805 22.65807 22.65807 22.03638 22.03619
#> 2  21.95658 21.95657 21.95656 21.95656 21.95656 21.95655 22.74101 22.74133
#> 3  17.65648 17.65644 17.65642 17.65640 17.65639 17.65638 17.22251 17.22244
#> 4  21.71092 21.71091 21.71090 21.71089 21.71088 21.71088 22.05297 22.05299
#> 5  25.01903 25.01905 25.01906 25.01906 25.01907 25.01907 24.10086 24.10025
#> 6  17.18397 17.18392 17.18389 17.18387 17.18386 17.18385 18.04496 18.04509
#> 7  17.84245 17.84241 17.84239 17.84237 17.84236 17.84235 18.44836 18.44836
#> 8  15.15927 15.15926 15.15926 15.15925 15.15925 15.15925 14.80664 14.80655
#> 9  16.72716 16.72715 16.72713 16.72713 16.72712 16.72712 15.92145 15.92112
#> 10 11.15682 11.15687 11.15691 11.15693 11.15695 11.15696 11.17600 11.17606
#> 11 10.43558 10.43564 10.43568 10.43571 10.43573 10.43574 10.64158 10.64174
#> 12 30.27942 30.27949 30.27953 30.27956 30.27958 30.27959 30.93072 30.93110
#> 13 31.89593 31.89598 31.89602 31.89604 31.89606 31.89607 32.11522 32.11537
#> 14 24.93405 24.93401 24.93398 24.93396 24.93395 24.93394 24.96410 24.96402
#> 15 17.24099 17.24095 17.24093 17.24092 17.24091 17.24090 16.92541 16.92540
#> 16 12.75791 12.75786 12.75782 12.75779 12.75778 12.75777 12.90704 12.90711
#> 17 16.41937 16.41936 16.41935 16.41934 16.41934 16.41933 16.24916 16.24923
#> 18 26.48893 26.48895 26.48896 26.48897 26.48898 26.48898 25.87605 25.87600
#> 19 18.96725 18.96728 18.96730 18.96731 18.96731 18.96732 19.61842 19.61888
#> 20 19.44180 19.44179 19.44179 19.44178 19.44178 19.44178 19.02026 19.02000
#> 21 14.06809 14.06809 14.06809 14.06809 14.06809 14.06810 14.20088 14.20077
#>      0.2076   0.1936   0.1806   0.1684   0.1571   0.1465   0.1366   0.1274
#> 1  22.86127 22.86155 22.86173 22.86186 22.86194 22.86198 22.86201 22.84540
#> 2  22.56755 22.56720 22.56691 22.56667 22.56648 22.56632 22.56619 22.99118
#> 3  16.98034 16.98039 16.98045 16.98052 16.98058 16.98065 16.98071 17.04358
#> 4  22.64387 22.64402 22.64410 22.64414 22.64414 22.64412 22.64409 22.61761
#> 5  23.36578 23.36617 23.36656 23.36693 23.36727 23.36759 23.36787 23.11543
#> 6  18.65743 18.65762 18.65772 18.65776 18.65775 18.65771 18.65766 18.47799
#> 7  18.89504 18.89530 18.89547 18.89558 18.89564 18.89567 18.89567 18.70961
#> 8  14.84359 14.84367 14.84375 14.84381 14.84387 14.84391 14.84395 14.90549
#> 9  15.66856 15.66884 15.66909 15.66932 15.66953 15.66971 15.66987 15.77194
#> 10 11.36447 11.36432 11.36419 11.36408 11.36398 11.36390 11.36384 11.37601
#> 11 10.79488 10.79461 10.79438 10.79418 10.79402 10.79389 10.79378 10.71817
#> 12 30.91748 30.91693 30.91648 30.91611 30.91581 30.91556 30.91536 31.03922
#> 13 31.84241 31.84207 31.84181 31.84162 31.84147 31.84136 31.84127 31.90393
#> 14 24.51227 24.51234 24.51242 24.51251 24.51259 24.51268 24.51276 24.54356
#> 15 17.14341 17.14349 17.14355 17.14359 17.14362 17.14363 17.14364 17.52665
#> 16 12.58732 12.58733 12.58734 12.58736 12.58738 12.58740 12.58743 12.17886
#> 17 16.17193 16.17182 16.17173 16.17167 16.17162 16.17159 16.17156 16.22660
#> 18 26.38411 26.38418 26.38421 26.38423 26.38423 26.38422 26.38420 25.97703
#> 19 18.89728 18.89663 18.89614 18.89575 18.89546 18.89524 18.89507 18.54128
#> 20 19.60362 19.60409 19.60444 19.60470 19.60489 19.60503 19.60512 19.91214
#> 21 13.29739 13.29744 13.29752 13.29762 13.29773 13.29785 13.29796 13.57830
#>      0.1188   0.1108   0.1033   0.0964   0.0899   0.0838   0.0782   0.0729
#> 1  22.84515 22.84498 22.84488 22.84483 22.84480 22.84480 22.84481 22.84483
#> 2  22.99150 22.99174 22.99192 22.99206 22.99217 22.99227 22.99234 22.99240
#> 3  17.04358 17.04355 17.04350 17.04344 17.04338 17.04332 17.04327 17.04322
#> 4  22.61749 22.61744 22.61743 22.61745 22.61750 22.61755 22.61760 22.61766
#> 5  23.11493 23.11450 23.11411 23.11378 23.11348 23.11321 23.11298 23.11277
#> 6  18.47789 18.47785 18.47787 18.47792 18.47800 18.47808 18.47816 18.47824
#> 7  18.70942 18.70930 18.70924 18.70921 18.70922 18.70924 18.70928 18.70931
#> 8  14.90542 14.90536 14.90531 14.90526 14.90522 14.90518 14.90515 14.90512
#> 9  15.77167 15.77142 15.77120 15.77099 15.77081 15.77066 15.77052 15.77040
#> 10 11.37609 11.37618 11.37627 11.37634 11.37641 11.37647 11.37652 11.37657
#> 11 10.71837 10.71856 10.71872 10.71887 10.71899 10.71910 10.71919 10.71926
#> 12 31.03966 31.04003 31.04033 31.04059 31.04080 31.04097 31.04111 31.04122
#> 13 31.90417 31.90435 31.90450 31.90461 31.90469 31.90474 31.90479 31.90482
#> 14 24.54350 24.54342 24.54333 24.54325 24.54317 24.54309 24.54302 24.54295
#> 15 17.52663 17.52660 17.52656 17.52653 17.52650 17.52649 17.52647 17.52647
#> 16 12.17894 12.17900 12.17904 12.17907 12.17908 12.17909 12.17908 12.17907
#> 17 16.22674 16.22684 16.22691 16.22696 16.22699 16.22702 16.22703 16.22704
#> 18 25.97702 25.97704 25.97709 25.97715 25.97721 25.97726 25.97730 25.97734
#> 19 18.54188 18.54236 18.54272 18.54300 18.54320 18.54335 18.54346 18.54353
#> 20 19.91174 19.91142 19.91116 19.91097 19.91083 19.91074 19.91067 19.91063
#> 21 13.57822 13.57807 13.57789 13.57772 13.57755 13.57739 13.57725 13.57714
#>      0.0680   0.0634   0.0591   0.0551   0.0514   0.0480   0.0447   0.0417
#> 1  22.84486 22.84488 22.84491 22.84494 22.84496 22.84498 22.84500 22.84502
#> 2  22.99246 22.99250 22.99254 22.99257 22.99259 22.99262 22.99264 22.99265
#> 3  17.04317 17.04313 17.04309 17.04306 17.04304 17.04302 17.04300 17.04298
#> 4  22.61771 22.61776 22.61781 22.61785 22.61789 22.61792 22.61795 22.61797
#> 5  23.11259 23.11244 23.11230 23.11218 23.11208 23.11200 23.11193 23.11186
#> 6  18.47832 18.47840 18.47846 18.47852 18.47857 18.47862 18.47866 18.47869
#> 7  18.70935 18.70939 18.70943 18.70946 18.70949 18.70952 18.70954 18.70956
#> 8  14.90510 14.90508 14.90507 14.90506 14.90505 14.90504 14.90503 14.90503
#> 9  15.77030 15.77022 15.77015 15.77009 15.77004 15.77000 15.76996 15.76994
#> 10 11.37660 11.37663 11.37666 11.37668 11.37669 11.37671 11.37672 11.37673
#> 11 10.71932 10.71937 10.71940 10.71943 10.71946 10.71948 10.71950 10.71951
#> 12 31.04131 31.04139 31.04145 31.04149 31.04153 31.04157 31.04159 31.04161
#> 13 31.90483 31.90485 31.90485 31.90486 31.90486 31.90486 31.90486 31.90485
#> 14 24.54289 24.54284 24.54280 24.54276 24.54273 24.54270 24.54268 24.54266
#> 15 17.52646 17.52647 17.52647 17.52648 17.52648 17.52649 17.52650 17.52651
#> 16 12.17906 12.17904 12.17903 12.17902 12.17900 12.17899 12.17898 12.17897
#> 17 16.22704 16.22704 16.22704 16.22704 16.22704 16.22704 16.22704 16.22703
#> 18 25.97738 25.97740 25.97743 25.97744 25.97746 25.97747 25.97748 25.97749
#> 19 18.54358 18.54362 18.54364 18.54365 18.54365 18.54365 18.54365 18.54365
#> 20 19.91061 19.91060 19.91060 19.91061 19.91062 19.91063 19.91064 19.91065
#> 21 13.57703 13.57694 13.57687 13.57681 13.57676 13.57671 13.57668 13.57665
#>      0.0389   0.0363   0.0338   0.0316   0.0294   0.0274   0.0256   0.0239
#> 1  22.84503 22.84505 22.84506 22.84507 22.84508 22.84508 22.84509 22.84509
#> 2  22.99266 22.99268 22.99268 22.99269 22.99270 22.99270 22.99271 22.99271
#> 3  17.04297 17.04296 17.04295 17.04294 17.04294 17.04293 17.04293 17.04292
#> 4  22.61799 22.61801 22.61803 22.61804 22.61805 22.61806 22.61806 22.61807
#> 5  23.11181 23.11177 23.11173 23.11170 23.11167 23.11165 23.11163 23.11162
#> 6  18.47872 18.47875 18.47877 18.47878 18.47880 18.47881 18.47882 18.47883
#> 7  18.70958 18.70959 18.70961 18.70962 18.70963 18.70963 18.70964 18.70965
#> 8  14.90502 14.90502 14.90502 14.90501 14.90501 14.90501 14.90501 14.90501
#> 9  15.76991 15.76989 15.76988 15.76986 15.76985 15.76984 15.76984 15.76983
#> 10 11.37673 11.37674 11.37675 11.37675 11.37675 11.37675 11.37676 11.37676
#> 11 10.71952 10.71953 10.71953 10.71954 10.71954 10.71954 10.71955 10.71955
#> 12 31.04163 31.04164 31.04165 31.04166 31.04167 31.04167 31.04168 31.04168
#> 13 31.90485 31.90485 31.90485 31.90484 31.90484 31.90484 31.90484 31.90484
#> 14 24.54264 24.54262 24.54261 24.54260 24.54259 24.54259 24.54258 24.54258
#> 15 17.52651 17.52652 17.52652 17.52653 17.52653 17.52654 17.52654 17.52654
#> 16 12.17896 12.17895 12.17895 12.17894 12.17893 12.17893 12.17893 12.17892
#> 17 16.22703 16.22703 16.22703 16.22703 16.22703 16.22703 16.22703 16.22703
#> 18 25.97749 25.97750 25.97750 25.97750 25.97750 25.97751 25.97751 25.97751
#> 19 18.54364 18.54364 18.54363 18.54363 18.54362 18.54362 18.54361 18.54361
#> 20 19.91066 19.91067 19.91068 19.91069 19.91070 19.91071 19.91071 19.91072
#> 21 13.57662 13.57660 13.57658 13.57657 13.57656 13.57655 13.57654 13.57654
#>      0.0223   0.0208   0.0194   0.0181   0.0168   0.0157   0.0146   0.0137
#> 1  22.84510 22.84510 22.84511 22.84511 22.84511 22.84511 22.84511 22.91232
#> 2  22.99271 22.99272 22.99272 22.99272 22.99272 22.99272 22.99272 22.97942
#> 3  17.04292 17.04292 17.04292 17.04291 17.04291 17.04291 17.04291 17.04086
#> 4  22.61808 22.61808 22.61808 22.61809 22.61809 22.61809 22.61809 22.56371
#> 5  23.11161 23.11160 23.11159 23.11158 23.11158 23.11157 23.11157 23.07221
#> 6  18.47884 18.47884 18.47885 18.47885 18.47885 18.47886 18.47886 18.47859
#> 7  18.70965 18.70965 18.70966 18.70966 18.70966 18.70966 18.70967 18.71001
#> 8  14.90501 14.90501 14.90501 14.90501 14.90501 14.90501 14.90501 14.91384
#> 9  15.76983 15.76982 15.76982 15.76982 15.76981 15.76981 15.76981 15.78351
#> 10 11.37676 11.37676 11.37676 11.37676 11.37676 11.37676 11.37676 11.38215
#> 11 10.71955 10.71955 10.71955 10.71955 10.71955 10.71955 10.71955 10.72572
#> 12 31.04168 31.04169 31.04169 31.04169 31.04169 31.04169 31.04169 31.06710
#> 13 31.90484 31.90483 31.90483 31.90483 31.90483 31.90483 31.90483 31.94557
#> 14 24.54257 24.54257 24.54257 24.54257 24.54256 24.54256 24.54256 24.58412
#> 15 17.52654 17.52655 17.52655 17.52655 17.52655 17.52655 17.52655 17.49998
#> 16 12.17892 12.17892 12.17892 12.17891 12.17891 12.17891 12.17891 12.25131
#> 17 16.22702 16.22702 16.22702 16.22702 16.22702 16.22702 16.22702 16.21435
#> 18 25.97751 25.97751 25.97751 25.97751 25.97751 25.97751 25.97751 25.92709
#> 19 18.54361 18.54360 18.54360 18.54360 18.54360 18.54360 18.54360 18.49217
#> 20 19.91072 19.91073 19.91073 19.91073 19.91073 19.91074 19.91074 19.88809
#> 21 13.57653 13.57653 13.57652 13.57652 13.57652 13.57652 13.57652 13.56788
#>      0.0127   0.0119   0.0111   0.0103   0.0096   0.0090   0.0084   0.0078
#> 1  22.91277 22.91312 22.91341 22.91372 22.95211 22.95186 22.95162 22.95140
#> 2  22.97901 22.97866 22.97838 22.97809 22.96909 22.96931 22.96950 22.96968
#> 3  17.04076 17.04069 17.04064 17.04066 17.06949 17.06972 17.06993 17.07010
#> 4  22.56352 22.56337 22.56323 22.56305 22.51457 22.51427 22.51401 22.51379
#> 5  23.07246 23.07269 23.07290 23.07306 23.09388 23.09411 23.09431 23.09448
#> 6  18.47869 18.47877 18.47881 18.47892 18.51849 18.51831 18.51814 18.51798
#> 7  18.71018 18.71032 18.71042 18.71058 18.76695 18.76687 18.76679 18.76670
#> 8  14.91385 14.91387 14.91389 14.91394 14.93245 14.93252 14.93258 14.93263
#> 9  15.78365 15.78378 15.78390 15.78405 15.82882 15.82908 15.82930 15.82948
#> 10 11.38215 11.38215 11.38214 11.38210 11.35862 11.35847 11.35835 11.35825
#> 11 10.72570 10.72567 10.72564 10.72558 10.68797 10.68775 10.68757 10.68742
#> 12 31.06696 31.06682 31.06670 31.06660 31.06394 31.06396 31.06398 31.06400
#> 13 31.94562 31.94566 31.94569 31.94575 31.97579 31.97594 31.97607 31.97618
#> 14 24.58426 24.58439 24.58450 24.58450 24.49857 24.49832 24.49813 24.49797
#> 15 17.49966 17.49941 17.49921 17.49903 17.51878 17.51901 17.51921 17.51938
#> 16 12.25167 12.25196 12.25219 12.25245 12.21678 12.21646 12.21619 12.21595
#> 17 16.21414 16.21398 16.21384 16.21377 16.22623 16.22638 16.22651 16.22662
#> 18 25.92701 25.92694 25.92687 25.92682 25.87236 25.87197 25.87164 25.87136
#> 19 18.49181 18.49151 18.49126 18.49114 18.53425 18.53464 18.53498 18.53528
#> 20 19.88816 19.88822 19.88827 19.88819 19.86239 19.86226 19.86215 19.86205
#> 21 13.56796 13.56803 13.56810 13.56800 13.53848 13.53877 13.53904 13.53928
#>      0.0073   0.0068   0.0063   0.0059   0.0055   0.0051
#> 1  22.95119 22.95099 22.95068 22.91838 22.91826 22.91815
#> 2  22.96985 22.97000 22.97022 22.95345 22.95341 22.95337
#> 3  17.07025 17.07038 17.07044 17.02203 17.02186 17.02170
#> 4  22.51361 22.51346 22.51319 22.51512 22.51516 22.51519
#> 5  23.09463 23.09475 23.09487 23.08155 23.08147 23.08140
#> 6  18.51782 18.51768 18.51761 18.52117 18.52126 18.52134
#> 7  18.76661 18.76651 18.76652 18.76015 18.76019 18.76023
#> 8  14.93267 14.93270 14.93322 15.02828 15.02861 15.02892
#> 9  15.82963 15.82975 15.83039 15.87639 15.87652 15.87664
#> 10 11.35817 11.35810 11.35762 11.31970 11.31958 11.31945
#> 11 10.68730 10.68720 10.68674 10.70004 10.70010 10.70016
#> 12 31.06402 31.06404 31.06387 31.10057 31.10073 31.10089
#> 13 31.97628 31.97635 31.97620 31.96128 31.96124 31.96120
#> 14 24.49785 24.49776 24.49777 24.50449 24.50439 24.50431
#> 15 17.51953 17.51966 17.52013 17.52419 17.52421 17.52424
#> 16 12.21575 12.21559 12.21522 12.21240 12.21233 12.21226
#> 17 16.22672 16.22681 16.22665 16.18147 16.18134 16.18121
#> 18 25.87113 25.87094 25.87021 25.88030 25.88037 25.88044
#> 19 18.53555 18.53578 18.53554 18.53762 18.53773 18.53784
#> 20 19.86196 19.86189 19.86236 19.85937 19.85930 19.85924
#> 21 13.53950 13.53970 13.54056 13.54205 13.54195 13.54184
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.4 18.7 24.4 22.8 19.2 17.8 16.4 15.2 10.4 10.4 32.4 33.9 21.5 15.5
#> [16] 13.3 19.2 26.0 15.8 19.7 15.0
#> 
#> $X
#>               am        carb        cyl       disp        drat       gear
#>  [1,]  1.4142136  0.55006532 -0.2236068 -0.7013330  0.67529395  0.3735437
#>  [2,] -0.7071068 -1.18264045 -0.2236068  0.1097619 -1.01736424 -0.9338592
#>  [3,] -0.7071068 -0.60507186  0.9503289  0.9539627 -0.87286903 -0.9338592
#>  [4,] -0.7071068 -0.60507186 -1.3975425 -0.8114102  0.24180831  0.3735437
#>  [5,] -0.7071068 -0.60507186 -1.3975425 -0.8602414  0.71657829  0.3735437
#>  [6,] -0.7071068  0.55006532 -0.2236068 -0.6384318  0.71657829  0.3735437
#>  [7,] -0.7071068  0.55006532 -0.2236068 -0.6384318  0.71657829  0.3735437
#>  [8,] -0.7071068 -0.02750327  0.9503289  0.2570832 -1.03800642 -0.9338592
#>  [9,] -0.7071068 -0.02750327  0.9503289  0.2570832 -1.03800642 -0.9338592
#> [10,] -0.7071068  0.55006532  0.9503289  1.8809283 -1.32699684 -0.9338592
#> [11,] -0.7071068  0.55006532  0.9503289  1.7816106 -1.18250163 -0.9338592
#> [12,]  1.4142136 -1.18264045 -1.3975425 -1.3742108  1.04685306  0.3735437
#> [13,]  1.4142136 -1.18264045 -1.3975425 -1.4371120  1.33584348  0.3735437
#> [14,] -0.7071068 -1.18264045 -1.3975425 -1.0315646  0.26245049 -0.9338592
#> [15,] -0.7071068 -0.60507186  0.9503289  0.6063506 -1.67791378 -0.9338592
#> [16,] -0.7071068  0.55006532  0.9503289  0.8711979  0.32437700 -0.9338592
#> [17,] -0.7071068 -0.60507186  0.9503289  1.2850219 -1.01736424 -0.9338592
#> [18,]  1.4142136 -0.60507186 -1.3975425 -1.0299093  1.76932912  1.6809466
#> [19,]  1.4142136  0.55006532  0.9503289  0.8794744  1.33584348  1.6809466
#> [20,]  1.4142136  1.70520250 -0.2236068 -0.8254802  0.09731310  1.6809466
#> [21,]  1.4142136  2.86033969  0.9503289  0.4656505 -0.06782428  1.6809466
#>               hp        qsec         vs          wt
#>  [1,] -0.6337194 -0.41541764 -0.7844645 -0.56398325
#>  [2,] -0.6337194  0.79998388  1.2747549 -0.17127402
#>  [3,]  0.2954181 -0.41541764 -0.7844645  0.08860708
#>  [4,] -1.3198517  1.08123382  1.2747549 -0.20014970
#>  [5,] -0.8481357  2.53770672  1.2747549 -0.24635079
#>  [6,] -0.4478919  0.22743936  1.2747549  0.08860708
#>  [7,] -0.4478919  0.52877858  1.2747549  0.08860708
#>  [8,]  0.3668902 -0.22456947 -0.7844645  0.81627418
#>  [9,]  0.3668902  0.07676975 -0.7844645  0.48131631
#> [10,]  0.7242507  0.06672511 -0.7844645  2.17920620
#> [11,]  0.8671950 -0.01363201 -0.7844645  2.38018092
#> [12,] -1.2626740  0.81505084  1.2747549 -1.34362657
#> [13,] -1.2769684  1.03101062  1.2747549 -1.76521147
#> [14,] -0.8195469  1.08625614  1.2747549 -1.03754438
#> [15,] -0.0619425 -0.49075244 -0.7844645  0.18100925
#> [16,]  1.2960276 -1.22401122 -0.7844645  0.55061794
#> [17,]  0.2954181 -0.40035068 -0.7844645  0.55639307
#> [18,] -0.9053134 -0.57613189 -0.7844645 -1.41292819
#> [19,]  1.5676216 -1.68104237 -0.7844645 -0.22325024
#> [20,]  0.2954181 -1.17881033 -0.7844645 -0.68526110
#> [21,]  2.5825256 -1.63081916 -0.7844645  0.23876061
#> attr(,"center")
#>  [1]   0.3333333   3.0476190   6.3809524 244.7380952   3.5728571   3.7142857
#>  [7] 154.3333333  17.8471429   0.3809524   3.3632857
#> attr(,"scale")
#>  [1]   0.4714045   1.7313961   1.7036708 120.8243293   0.4844451   0.7648752
#>  [7]  69.9573566   1.9911115   0.4856209   0.8657805
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 40
#> 
#> $lambda.min
#> [1] 0.338381
#> 
#> $null.dev
#> [1] 35.55324
#> 
#> $Bias
#> [1] 3.123985
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 11.37668