Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 30.170761 27.785453 24.077987 20.354695 17.263761 14.712887 12.622464
#>   [8] 10.923878  9.558035  8.474079  7.628287  6.983100  6.506296  6.170256
#>  [15]  5.951345  5.829363  5.637980  5.464995  5.458944  5.458944  5.458944
#>  [22]  5.458944  5.458944  5.458944  5.458944  5.458944  5.458944  5.458944
#>  [29]  5.649739  5.949809  6.554469  7.352401  7.239807  6.659427  8.645293
#>  [36]  8.043643  8.028259  8.015498  7.976650  7.993886 10.298038 11.235558
#>  [43] 11.292189 11.380052 11.455345 11.183069 11.208181 11.191824 11.591874
#>  [50] 11.616068 11.680181 11.672574 13.165632 13.267202 13.190573 11.869831
#>  [57] 11.868973 11.767892 11.277679 11.477988 11.478192 11.638989 11.638709
#>  [64] 11.942557 11.913237 11.913435 11.736306 11.391126 11.391105 12.767734
#>  [71] 12.769655 12.400372 12.391470 12.106583 12.257096 12.326964 12.271931
#>  [78] 12.381542 12.381495 12.396947 12.415474 12.563780 12.563460 12.576186
#>  [85] 12.575869 12.575567 12.499928 12.508399 13.016645 13.017112 13.017583
#>  [92] 13.018098 13.114361 13.115011 13.126891 13.127440 13.127982 13.137557
#>  [99] 13.138090 13.138617
#> 
#> $cvse
#>   [1] 7.705806 7.632594 6.895849 6.034688 5.300753 4.673430 4.135566 3.673137
#>   [9] 3.274927 2.932221 2.638492 2.389077 2.180815 2.011628 1.880015 1.784495
#>  [17] 1.682085 1.587888 1.587991 1.587991 1.587991 1.587991 1.587991 1.587991
#>  [25] 1.587991 1.587991 1.587991 1.587991 1.606958 1.635951 1.709700 1.710104
#>  [33] 1.692397 1.728300 2.402023 2.379100 2.380028 2.381817 2.376279 2.373129
#>  [41] 4.124545 4.220848 4.218123 4.214011 4.211141 4.158508 4.177818 4.183616
#>  [49] 4.296307 4.293449 4.297027 4.297416 4.440116 4.479317 4.489993 3.270414
#>  [57] 3.269791 3.304875 3.273363 3.353885 3.353491 3.357232 3.357028 3.353361
#>  [65] 3.346930 3.346827 3.068995 2.934530 2.934374 3.311370 3.321077 3.205707
#>  [73] 3.206331 3.088160 3.125655 3.134120 3.139491 3.121600 3.121568 3.120766
#>  [81] 3.122576 3.121824 3.121736 3.119837 3.119764 3.119696 3.079743 3.078370
#>  [89] 3.348799 3.349070 3.349342 3.349629 3.376911 3.377227 3.375456 3.375781
#>  [97] 3.376102 3.377698 3.378013 3.378325
#> 
#> $fold
#>  [1]  9  4  8  6  4  9  2 10  5  1  2  5  7  8  7  1  3  1  6 10  3
#> 
#> $lambda
#>   [1] 4.941721986 4.608666464 4.298057769 4.008383060 3.738231457 3.486287169
#>   [7] 3.251323082 3.032194788 2.827835007 2.637248392 2.459506677 2.293744159
#>  [13] 2.139153480 1.994981695 1.860526605 1.735133339 1.618191159 1.509130491
#>  [19] 1.407420147 1.312564739 1.224102269 1.141601873 1.064661728 0.992907090
#>  [25] 0.925988476 0.863579952 0.805377553 0.751097802 0.700476324 0.653266564
#>  [31] 0.609238584 0.568177942 0.529884651 0.494172199 0.460866646 0.429805777
#>  [37] 0.400838306 0.373823146 0.348628717 0.325132308 0.303219479 0.282783501
#>  [43] 0.263724839 0.245950668 0.229374416 0.213915348 0.199498169 0.186052661
#>  [49] 0.173513334 0.161819116 0.150913049 0.140742015 0.131256474 0.122410227
#>  [55] 0.114160187 0.106466173 0.099290709 0.092598848 0.086357995 0.080537755
#>  [61] 0.075109780 0.070047632 0.065326656 0.060923858 0.056817794 0.052988465
#>  [67] 0.049417220 0.046086665 0.042980578 0.040083831 0.037382315 0.034862872
#>  [73] 0.032513231 0.030321948 0.028278350 0.026372484 0.024595067 0.022937442
#>  [79] 0.021391535 0.019949817 0.018605266 0.017351333 0.016181912 0.015091305
#>  [85] 0.014074201 0.013125647 0.012241023 0.011416019 0.010646617 0.009929071
#>  [91] 0.009259885 0.008635800 0.008053776 0.007510978 0.007004763 0.006532666
#>  [97] 0.006092386 0.005681779 0.005298847 0.004941722
#> 
#> $fit
#> $beta
#>               4.9417       4.6087       4.2981      4.0084      3.7382
#> (Intercept) 18.88571 19.868471752 20.784994654 21.63974698 22.43689186
#> am           0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> carb         0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> cyl          0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> disp         0.00000 -0.004031786 -0.007791843 -0.01129849 -0.01456879
#> drat         0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> gear         0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> hp           0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> qsec         0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> vs           0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> wt           0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#>                  3.4863      3.2513      3.0322      2.8278      2.6372
#> (Intercept) 23.18031184 23.87362781 24.52021660 25.12322747 25.68559742
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.01761869 -0.02046304 -0.02311568 -0.02558955 -0.02789668
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  2.4595      2.2937      2.1392      1.9950      1.8605
#> (Intercept) 26.21006552 26.69918623 27.15534184 27.58075408 27.97749497
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.03004833 -0.03205496 -0.03392635 -0.03567161 -0.03729925
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  1.7351      1.6182      1.5091      1.4074      1.3126
#> (Intercept) 28.34749685 28.60684455 28.60684455 28.60684455 28.60684455
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.03881719 -0.03988117 -0.03988117 -0.03988117 -0.03988117
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  1.2241      1.1416      1.0647      0.9929      0.9260
#> (Intercept) 28.60684455 28.60684455 28.60684455 28.60684455 28.60684455
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.03988117 -0.03988117 -0.03988117 -0.03988117 -0.03988117
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.8636      0.8054      0.7511      0.7005      0.6533
#> (Intercept) 28.60684455 28.60684455 28.60684455 28.60684455 28.60684455
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.03988117 -0.03988117 -0.03988117 -0.03988117 -0.03988117
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.6092      0.5682      0.5299      0.4942      0.4609
#> (Intercept) 28.60684455 28.60684455 26.86376442 16.81337780 16.81238115
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.03988117 -0.03988117 -0.03853535 -0.01887179 -0.01886890
#> drat         0.00000000  0.00000000  0.40069688  2.91282294  2.91303716
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000 -0.02418578 -0.02418888
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.4298      0.4008      0.3738      0.3486      0.3251
#> (Intercept) 16.81163592 16.81107870 16.81066205 16.81035051 16.81011757
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.01886674 -0.01886512 -0.01886391 -0.01886301 -0.01886233
#> drat         2.91319734  2.91331711  2.91340667  2.91347363  2.91352370
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.02419121 -0.02419295 -0.02419425 -0.02419522 -0.02419595
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.3032      0.2828      0.2637      0.2460      0.2294
#> (Intercept) 16.80994339 16.80981315 17.08073693 17.51756621 17.92300616
#> am           0.00000000  0.00000000  0.23990400  0.62080998  0.97476559
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.01886183 -0.01886145 -0.01839611 -0.01766941 -0.01699347
#> drat         2.91356114  2.91358913  2.81373544  2.65368495  2.50506989
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.02419649 -0.02419689 -0.02486792 -0.02592252 -0.02690298
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.2139      0.1995      0.1861      0.1735      0.1618
#> (Intercept) 18.30280469 18.40154017 18.40373400 18.40528346 18.40636118
#> am           1.30582713  1.39435665  1.39488004  1.39514861  1.39525649
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.01636223 -0.01618586 -0.01618765 -0.01618951 -0.01619127
#> drat         2.36593483  2.32934265  2.32876129  2.32836895  2.32811032
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.02781919 -0.02807214 -0.02807116 -0.02806973 -0.02806817
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.1509      0.1407      0.1313      0.1224      0.1142
#> (Intercept) 12.25849667 12.25590546 12.25390240 12.25235526 12.25116105
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         2.23624562  2.23653712  2.23676285  2.23693745  2.23707237
#> gear         1.14221666  1.14204352  1.14190992  1.14180688  1.14172743
#> hp          -0.02606372 -0.02605879 -0.02605499 -0.02605206 -0.02604981
#> qsec         0.29587837  0.29598359  0.29606476  0.29612735  0.29617561
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -1.99716164 -1.99730135 -1.99740872 -1.99749126 -1.99755474
#>                  0.1065      0.0993      0.0926     0.0864      0.0805
#> (Intercept) 12.25023975 12.24952928 12.24898157 12.2485594 12.24823418
#> am           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> drat         2.23717655  2.23725694  2.23731895  2.2373668  2.23740361
#> gear         1.14166620  1.14161901  1.14158266  1.1415546  1.14153308
#> hp          -0.02604807 -0.02604673 -0.02604569 -0.0260449 -0.02604429
#> qsec         0.29621280  0.29624145  0.29626353  0.2962805  0.29629363
#> vs           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> wt          -1.99760357 -1.99764114 -1.99767004 -1.9976923 -1.99770941
#>                  0.0751      0.0700      0.0653      0.0609      0.0568
#> (Intercept) 12.24798359 12.24779055 11.98098293 11.97875264 11.97762969
#> am           0.00000000  0.00000000  0.56029556  0.55989231  0.55951522
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         2.23743201  2.23745389  2.16487912  2.16540339  2.16571569
#> gear         1.14151646  1.14150366  0.95604874  0.95610401  0.95619791
#> hp          -0.02604382 -0.02604345 -0.02577551 -0.02577515 -0.02577534
#> qsec         0.29630372  0.29631149  0.33744816  0.33744362  0.33742407
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -1.99772259 -1.99773273 -1.92984995 -1.92975321 -1.92970015
#>                  0.0530      0.0494      0.0461     0.0430      0.0401
#> (Intercept) 11.97721619 11.97722890 11.97747374 11.9778234 11.97819881
#> am           0.55918939  0.55892299  0.55871444  0.5585572  0.55844260
#> carb         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> drat         2.16588365  2.16595759  2.16597365  2.1659568  2.16592372
#> gear         0.95630382  0.95640629  0.95649734  0.9565738  0.95663522
#> hp          -0.02577583 -0.02577645 -0.02577709 -0.0257777 -0.02577823
#> qsec         0.33739797  0.33737054  0.33734479  0.3373222  0.33730333
#> vs           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> wt          -1.92967455 -1.92966533 -1.92966518 -1.9296694 -1.92967531
#>                  0.0374      0.0349      0.0325      0.0303      0.0283
#> (Intercept) 11.04014786 11.03881567 11.03752920 11.03628733 11.03508898
#> am           0.62576182  0.62585857  0.62594773  0.62602977  0.62610514
#> carb         0.31701219  0.31707959  0.31714099  0.31719679  0.31724736
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.01008914  0.01009326  0.01009719  0.01010095  0.01010453
#> drat         2.06192986  2.06188628  2.06184641  2.06181000  2.06177684
#> gear         0.99556959  0.99559221  0.99561745  0.99564502  0.99567465
#> hp          -0.03365284 -0.03365355 -0.03365418 -0.03365473 -0.03365519
#> qsec         0.43760359  0.43769793  0.43778751  0.43787254  0.43795324
#> vs           0.39534282  0.39538200  0.39542147  0.39546110  0.39550080
#> wt          -2.81552031 -2.81594258 -2.81634468 -2.81672748 -2.81709182
#>                  0.0264      0.0246      0.0229      0.0214      0.0199
#> (Intercept) 11.03393302 11.03281836 11.03174389 11.03070851 11.02971113
#> am           0.62617425  0.62623752  0.62629529  0.62634795  0.62639581
#> carb         0.31729305  0.31733419  0.31737108  0.31740402  0.31743329
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.01010796  0.01011122  0.01011433  0.01011730  0.01012013
#> drat         2.06174672  2.06171943  2.06169478  2.06167259  2.06165270
#> gear         0.99570608  0.99573909  0.99577345  0.99580898  0.99584547
#> hp          -0.03365559 -0.03365592 -0.03365618 -0.03365639 -0.03365655
#> qsec         0.43802979  0.43810240  0.43817125  0.43823651  0.43829835
#> vs           0.39554045  0.39557999  0.39561932  0.39565837  0.39569708
#> wt          -2.81743850 -2.81776829 -2.81808193 -2.81838014 -2.81866360
#>                  0.0186      0.0174      0.0162      0.0151      0.0141
#> (Intercept) 11.02875067 11.02782606 11.02693624 11.02608016 11.02525680
#> am           0.62643920  0.62647841  0.62651372  0.62654540  0.62657371
#> carb         0.31745914  0.31748182  0.31750157  0.31751858  0.31753308
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.01012282  0.01012539  0.01012784  0.01013017  0.01013239
#> drat         2.06163495  2.06161919  2.06160527  2.06159307  2.06158246
#> gear         0.99588277  0.99592071  0.99595915  0.99599795  0.99603701
#> hp          -0.03365666 -0.03365673 -0.03365675 -0.03365674 -0.03365670
#> qsec         0.43835693  0.43841242  0.43846495  0.43851467  0.43856171
#> vs           0.39573539  0.39577324  0.39581060  0.39584741  0.39588365
#> wt          -2.81893297 -2.81918890 -2.81943197 -2.81966278 -2.81988189
#>                  0.0131      0.0122      0.0114      0.0106      0.0099
#> (Intercept) 11.02446515 11.02370420 11.02297298 11.02227054 11.02159592
#> am           0.62659886  0.62662110  0.62664062  0.62665762  0.62667230
#> carb         0.31754525  0.31755527  0.31756330  0.31756951  0.31757405
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.01013450  0.01013652  0.01013843  0.01014025  0.01014198
#> drat         2.06157333  2.06156555  2.06155904  2.06155369  2.06154941
#> gear         0.99607619  0.99611540  0.99615455  0.99619356  0.99623234
#> hp          -0.03365662 -0.03365652 -0.03365640 -0.03365625 -0.03365608
#> qsec         0.43860620  0.43864828  0.43868805  0.43872562  0.43876112
#> vs           0.39591928  0.39595426  0.39598859  0.39602223  0.39605517
#> wt          -2.82008983 -2.82028711 -2.82047424 -2.82065168 -2.82081988
#>                  0.0093      0.0086      0.0081      0.0075      0.0070
#> (Intercept) 11.02094821 11.02032650 11.01972992 11.01915761 11.01860872
#> am           0.62668481  0.62669533  0.62670400  0.62671098  0.62671639
#> carb         0.31757704  0.31757863  0.31757894  0.31757807  0.31757614
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.01014362  0.01014519  0.01014667  0.01014808  0.01014943
#> drat         2.06154612  2.06154373  2.06154217  2.06154138  2.06154127
#> gear         0.99627082  0.99630895  0.99634667  0.99638392  0.99642067
#> hp          -0.03365590 -0.03365569 -0.03365548 -0.03365525 -0.03365501
#> qsec         0.43879464  0.43882628  0.43885613  0.43888429  0.43891084
#> vs           0.39608741  0.39611891  0.39614969  0.39617973  0.39620902
#> wt          -2.82097929 -2.82113031 -2.82127334 -2.82140878 -2.82153697
#>                  0.0065      0.0061      0.0057      0.0053      0.0049
#> (Intercept) 11.01808243 11.01757796 11.01709451 11.01663133 11.01618769
#> am           0.62672037  0.62672303  0.62672449  0.62672484  0.62672419
#> carb         0.31757324  0.31756947  0.31756491  0.31755964  0.31755374
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.01015070  0.01015191  0.01015305  0.01015414  0.01015517
#> drat         2.06154180  2.06154291  2.06154454  2.06154663  2.06154916
#> gear         0.99645687  0.99649249  0.99652749  0.99656186  0.99659557
#> hp          -0.03365477 -0.03365451 -0.03365425 -0.03365399 -0.03365372
#> qsec         0.43893586  0.43895943  0.43898163  0.43900252  0.43902218
#> vs           0.39623758  0.39626539  0.39629246  0.39631879  0.39634439
#> wt          -2.82165827 -2.82177301 -2.82188151 -2.82198409 -2.82208102
#> 
#> $iter
#>   [1]   0   4   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1   1   1   1   1   1   1   1   1   1  23  69   1   1
#>  [37]   1   1   1   1   1   1  25  27  26  26   9   1   1   1 110   1   1   1
#>  [55]   1   1   1   1   1   1   1   1  26   1   1   1   1   1   1   1 180   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.941721986 4.608666464 4.298057769 4.008383060 3.738231457 3.486287169
#>   [7] 3.251323082 3.032194788 2.827835007 2.637248392 2.459506677 2.293744159
#>  [13] 2.139153480 1.994981695 1.860526605 1.735133339 1.618191159 1.509130491
#>  [19] 1.407420147 1.312564739 1.224102269 1.141601873 1.064661728 0.992907090
#>  [25] 0.925988476 0.863579952 0.805377553 0.751097802 0.700476324 0.653266564
#>  [31] 0.609238584 0.568177942 0.529884651 0.494172199 0.460866646 0.429805777
#>  [37] 0.400838306 0.373823146 0.348628717 0.325132308 0.303219479 0.282783501
#>  [43] 0.263724839 0.245950668 0.229374416 0.213915348 0.199498169 0.186052661
#>  [49] 0.173513334 0.161819116 0.150913049 0.140742015 0.131256474 0.122410227
#>  [55] 0.114160187 0.106466173 0.099290709 0.092598848 0.086357995 0.080537755
#>  [61] 0.075109780 0.070047632 0.065326656 0.060923858 0.056817794 0.052988465
#>  [67] 0.049417220 0.046086665 0.042980578 0.040083831 0.037382315 0.034862872
#>  [73] 0.032513231 0.030321948 0.028278350 0.026372484 0.024595067 0.022937442
#>  [79] 0.021391535 0.019949817 0.018605266 0.017351333 0.016181912 0.015091305
#>  [85] 0.014074201 0.013125647 0.012241023 0.011416019 0.010646617 0.009929071
#>  [91] 0.009259885 0.008635800 0.008053776 0.007510978 0.007004763 0.006532666
#>  [97] 0.006092386 0.005681779 0.005298847 0.004941722
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 32
#> 
#> $loss
#>   [1] 605.24571 506.79730 424.43055 355.83121 299.00123 252.21662 213.99082
#>   [8] 183.04290 158.26997 138.72322 123.58719 112.16170 103.84625  98.12643
#>  [15]  94.56222  92.77779  92.41277  92.41277  92.41277  92.41277  92.41277
#>  [22]  92.41277  92.41277  92.41277  92.41277  92.41277  92.41277  92.41277
#>  [29]  92.41277  92.41277  92.41277  92.41277  87.60149  55.43544  55.43543
#>  [36]  55.43543  55.43543  55.43543  55.43543  55.43543  55.43543  55.43543
#>  [43]  54.29049  52.91531  52.12436  51.80913  51.79438  51.79438  51.79437
#>  [50]  51.79437  42.87149  42.87148  42.87148  42.87148  42.87148  42.87148
#>  [57]  42.87148  42.87147  42.87147  42.87147  42.87147  42.87147  42.50325
#>  [64]  42.50324  42.50324  42.50324  42.50324  42.50324  42.50324  42.50324
#>  [71]  41.01558  41.01557  41.01556  41.01556  41.01555  41.01554  41.01553
#>  [78]  41.01553  41.01552  41.01552  41.01552  41.01551  41.01551  41.01550
#>  [85]  41.01550  41.01550  41.01550  41.01550  41.01549  41.01549  41.01549
#>  [92]  41.01549  41.01549  41.01549  41.01549  41.01548  41.01548  41.01548
#>  [99]  41.01548  41.01548
#> 
#> $linear.predictors
#>      4.9417   4.6087   4.2981   4.0084   3.7382   3.4863   3.2513   3.0322
#> 1  18.88571 19.43304 19.94348 20.41951 20.86346 21.27749 21.66362 22.02372
#> 2  18.88571 18.82827 18.77470 18.72474 18.67814 18.63469 18.59416 18.55637
#> 3  18.88571 18.96132 19.03183 19.09759 19.15891 19.21611 19.26944 19.31919
#> 4  18.88571 18.41703 17.97993 17.57229 17.19213 16.83758 16.50694 16.19857
#> 5  18.88571 19.27701 19.64193 19.98226 20.29965 20.59565 20.87170 21.12915
#> 6  18.88571 19.30080 19.68790 20.04892 20.38561 20.69960 20.99243 21.26553
#> 7  18.88571 19.19274 19.47908 19.74612 19.99516 20.22742 20.44402 20.64603
#> 8  18.88571 18.75651 18.63600 18.52362 18.41882 18.32108 18.22992 18.14491
#> 9  18.88571 18.75651 18.63600 18.52362 18.41882 18.32108 18.22992 18.14491
#> 10 18.88571 17.96547 17.10724 16.30686 15.56042 14.86429 14.21508 13.60962
#> 11 18.88571 18.01385 17.20075 16.44244 15.73525 15.07571 14.46063 13.88700
#> 12 18.88571 18.09449 17.35658 16.66841 16.02662 15.42809 14.86989 14.34932
#> 13 18.88571 19.56327 20.19515 20.78445 21.33403 21.84658 22.32458 22.77036
#> 14 18.88571 19.38425 19.84919 20.28280 20.68718 21.06431 21.41602 21.74402
#> 15 18.88571 18.58636 18.30719 18.04683 17.80402 17.57757 17.36638 17.16943
#> 16 18.88571 18.64281 18.41627 18.20501 18.00798 17.82423 17.65287 17.49305
#> 17 18.88571 18.45735 18.05785 17.68528 17.33781 17.01377 16.71157 16.42973
#> 18 18.88571 19.54996 20.16944 20.74717 21.28596 21.78844 22.25705 22.69408
#> 19 18.88571 19.38345 19.84764 20.28054 20.68427 21.06078 21.41192 21.73940
#> 20 18.88571 18.65490 18.43965 18.23890 18.05169 17.87709 17.71425 17.56240
#> 21 18.88571 19.38063 19.84218 20.27263 20.67407 21.04845 21.39760 21.72322
#>      2.8278   2.6372   2.4595   2.2937   2.1392   1.9950   1.8605   1.7351
#> 1  22.35956 22.67276 22.96485 23.23725 23.49130 23.72822 23.94918 24.15524
#> 2  18.52112 18.48825 18.45760 18.42901 18.40234 18.37748 18.35429 18.33266
#> 3  19.36558 19.40884 19.44919 19.48682 19.52191 19.55464 19.58516 19.61363
#> 4  15.91099 15.64279 15.39267 15.15940 14.94186 14.73897 14.54977 14.37331
#> 5  21.36924 21.59315 21.80198 21.99672 22.17835 22.34773 22.50570 22.65302
#> 6  21.52022 21.75774 21.97926 22.18585 22.37851 22.55819 22.72576 22.88204
#> 7  20.83442 21.01011 21.17397 21.32678 21.46929 21.60219 21.72614 21.84174
#> 8  18.06563 17.99169 17.92274 17.85843 17.79846 17.74252 17.69036 17.64172
#> 9  18.06563 17.99169 17.92274 17.85843 17.79846 17.74252 17.69036 17.64172
#> 10 13.04496 12.51836 12.02726 11.56925 11.14211 10.74375 10.37225 10.02578
#> 11 13.35204 12.85312 12.38784 11.95391 11.54922 11.17181 10.81984 10.49159
#> 12 13.86383 13.41106 12.98880 12.59501 12.22775 11.88525 11.56583 11.26793
#> 13 23.18610 23.57382 23.93541 24.27263 24.58712 24.88041 25.15394 25.40904
#> 14 22.04992 22.33521 22.60126 22.84939 23.08079 23.29659 23.49786 23.68555
#> 15 16.98575 16.81445 16.65470 16.50571 16.36676 16.23718 16.11633 16.00363
#> 16 17.34400 17.20501 17.07537 16.95448 16.84173 16.73658 16.63852 16.54707
#> 17 16.16689 15.92176 15.69315 15.47995 15.28112 15.09569 14.92276 14.76148
#> 18 23.10165 23.48176 23.83625 24.16684 24.47516 24.76270 25.03085 25.28094
#> 19 22.04480 22.32963 22.59525 22.84297 23.07400 23.28946 23.49040 23.67779
#> 20 17.42077 17.28870 17.16552 17.05064 16.94351 16.84360 16.75042 16.66352
#> 21 22.02689 22.31010 22.57422 22.82054 23.05025 23.26449 23.46429 23.65062
#>       1.6182    1.5091    1.4074    1.3126    1.2241    1.1416    1.0647
#> 1  24.299678 24.299678 24.299678 24.299678 24.299678 24.299678 24.299678
#> 2  18.317503 18.317503 18.317503 18.317503 18.317503 18.317503 18.317503
#> 3  19.633581 19.633581 19.633581 19.633581 19.633581 19.633581 19.633581
#> 4  14.249623 14.249623 14.249623 14.249623 14.249623 14.249623 14.249623
#> 5  22.756277 22.756277 22.756277 22.756277 22.756277 22.756277 22.756277
#> 6  22.991576 22.991576 22.991576 22.991576 22.991576 22.991576 22.991576
#> 7  21.922760 21.922760 21.922760 21.922760 21.922760 21.922760 21.922760
#> 8  17.607618 17.607618 17.607618 17.607618 17.607618 17.607618 17.607618
#> 9  17.607618 17.607618 17.607618 17.607618 17.607618 17.607618 17.607618
#> 10  9.782932  9.782932  9.782932  9.782932  9.782932  9.782932  9.782932
#> 11 10.261506 10.261506 10.261506 10.261506 10.261506 10.261506 10.261506
#> 12 11.059130 11.059130 11.059130 11.059130 11.059130 11.059130 11.059130
#> 13 25.587840 25.587840 25.587840 25.587840 25.587840 25.587840 25.587840
#> 14 23.817116 23.817116 23.817116 23.817116 23.817116 23.817116 23.817116
#> 15 15.924632 15.924632 15.924632 15.924632 15.924632 15.924632 15.924632
#> 16 16.482969 16.482969 16.482969 16.482969 16.482969 16.482969 16.482969
#> 17 14.648435 14.648435 14.648435 14.648435 14.648435 14.648435 14.648435
#> 18 25.456232 25.456232 25.456232 25.456232 25.456232 25.456232 25.456232
#> 19 23.809140 23.809140 23.809140 23.809140 23.809140 23.809140 23.809140
#> 20 16.602612 16.602612 16.602612 16.602612 16.602612 16.602612 16.602612
#> 21 23.781223 23.781223 23.781223 23.781223 23.781223 23.781223 23.781223
#>       0.9929    0.9260    0.8636    0.8054    0.7511    0.7005    0.6533
#> 1  24.299678 24.299678 24.299678 24.299678 24.299678 24.299678 24.299678
#> 2  18.317503 18.317503 18.317503 18.317503 18.317503 18.317503 18.317503
#> 3  19.633581 19.633581 19.633581 19.633581 19.633581 19.633581 19.633581
#> 4  14.249623 14.249623 14.249623 14.249623 14.249623 14.249623 14.249623
#> 5  22.756277 22.756277 22.756277 22.756277 22.756277 22.756277 22.756277
#> 6  22.991576 22.991576 22.991576 22.991576 22.991576 22.991576 22.991576
#> 7  21.922760 21.922760 21.922760 21.922760 21.922760 21.922760 21.922760
#> 8  17.607618 17.607618 17.607618 17.607618 17.607618 17.607618 17.607618
#> 9  17.607618 17.607618 17.607618 17.607618 17.607618 17.607618 17.607618
#> 10  9.782932  9.782932  9.782932  9.782932  9.782932  9.782932  9.782932
#> 11 10.261506 10.261506 10.261506 10.261506 10.261506 10.261506 10.261506
#> 12 11.059130 11.059130 11.059130 11.059130 11.059130 11.059130 11.059130
#> 13 25.587840 25.587840 25.587840 25.587840 25.587840 25.587840 25.587840
#> 14 23.817116 23.817116 23.817116 23.817116 23.817116 23.817116 23.817116
#> 15 15.924632 15.924632 15.924632 15.924632 15.924632 15.924632 15.924632
#> 16 16.482969 16.482969 16.482969 16.482969 16.482969 16.482969 16.482969
#> 17 14.648435 14.648435 14.648435 14.648435 14.648435 14.648435 14.648435
#> 18 25.456232 25.456232 25.456232 25.456232 25.456232 25.456232 25.456232
#> 19 23.809140 23.809140 23.809140 23.809140 23.809140 23.809140 23.809140
#> 20 16.602612 16.602612 16.602612 16.602612 16.602612 16.602612 16.602612
#> 21 23.781223 23.781223 23.781223 23.781223 23.781223 23.781223 23.781223
#>       0.6092    0.5682    0.5299   0.4942   0.4609   0.4298   0.4008   0.3738
#> 1  24.299678 24.299678 24.244630 23.74032 23.74017 23.74006 23.73997 23.73991
#> 2  18.317503 18.317503 18.155791 18.25552 18.25558 18.25563 18.25567 18.25570
#> 3  19.633581 19.633581 19.299235 18.06711 18.06703 18.06697 18.06692 18.06689
#> 4  14.249623 14.249623 14.277277 13.44418 13.44415 13.44413 13.44411 13.44410
#> 5  22.756277 22.756277 22.689201 23.29368 23.29371 23.29373 23.29374 23.29375
#> 6  22.991576 22.991576 23.008719 23.27685 23.27680 23.27677 23.27674 23.27672
#> 7  21.922760 21.922760 21.975972 22.09388 22.09383 22.09379 22.09375 22.09373
#> 8  17.607618 17.607618 17.465855 16.19746 16.19736 16.19729 16.19723 16.19719
#> 9  17.607618 17.607618 17.465855 16.19746 16.19736 16.19729 16.19723 16.19719
#> 10  9.782932  9.782932  9.849123 11.48238 11.48274 11.48301 11.48321 11.48336
#> 11 10.261506 10.261506 10.339596 11.67088 11.67119 11.67142 11.67159 11.67172
#> 12 11.059130 11.059130 11.202463 12.35548 12.35573 12.35592 12.35606 12.35617
#> 13 25.587840 25.587840 25.922074 28.48734 28.48746 28.48754 28.48761 28.48766
#> 14 23.817116 23.817116 23.718248 22.97830 22.97814 22.97802 22.97794 22.97787
#> 15 15.924632 15.924632 15.715448 15.22367 15.22372 15.22376 15.22378 15.22380
#> 16 16.482969 16.482969 16.411214 16.62388 16.62397 16.62404 16.62409 16.62413
#> 17 14.648435 14.648435 14.870992 15.14757 15.14762 15.14766 15.14769 15.14771
#> 18 25.456232 25.456232 25.454315 25.61056 25.61046 25.61039 25.61033 25.61029
#> 19 23.809140 23.809140 24.003049 25.24600 25.24602 25.24603 25.24604 25.24605
#> 20 16.602612 16.602612 16.683092 13.34213 13.34172 13.34141 13.34118 13.34101
#> 21 23.781223 23.781223 23.847852 23.86534 23.86524 23.86516 23.86510 23.86506
#>      0.3486   0.3251   0.3032   0.2828   0.2637   0.2460   0.2294   0.2139
#> 1  23.73986 23.73983 23.73980 23.73978 23.85403 24.03597 24.20502 24.36318
#> 2  18.25572 18.25573 18.25575 18.25576 18.26538 18.28073 18.29498 18.30832
#> 3  18.06686 18.06684 18.06683 18.06682 18.09639 18.14426 18.18866 18.23027
#> 4  13.44409 13.44408 13.44408 13.44407 13.39759 13.32389 13.25540 13.19135
#> 5  23.29376 23.29377 23.29377 23.29377 23.22290 23.11037 23.00579 22.90798
#> 6  23.27671 23.27670 23.27669 23.27669 23.15796 22.96952 22.79442 22.63064
#> 7  22.09371 22.09370 22.09369 22.09369 21.96864 21.77015 21.58571 21.41320
#> 8  16.19716 16.19713 16.19712 16.19710 16.16903 16.12510 16.08424 16.04607
#> 9  16.19716 16.19713 16.19712 16.19710 16.16903 16.12510 16.08424 16.04607
#> 10 11.48347 11.48355 11.48361 11.48366 11.54410 11.63879 11.72683 11.80909
#> 11 11.67182 11.67189 11.67194 11.67198 11.71313 11.77735 11.83708 11.89286
#> 12 12.35625 12.35631 12.35635 12.35638 12.35519 12.35225 12.34957 12.34698
#> 13 28.48769 28.48772 28.48774 28.48776 28.50664 28.53550 28.56241 28.58747
#> 14 22.97782 22.97778 22.97775 22.97773 22.87000 22.69962 22.54126 22.39320
#> 15 15.22382 15.22383 15.22384 15.22384 15.26650 15.33449 15.39763 15.45672
#> 16 16.62416 16.62418 16.62419 16.62420 16.62140 16.61680 16.61252 16.60850
#> 17 15.14773 15.14774 15.14775 15.14775 15.04469 14.88050 14.72797 14.58526
#> 18 25.61026 25.61024 25.61022 25.61021 25.70611 25.85864 26.00038 26.13296
#> 19 25.24605 25.24606 25.24606 25.24606 25.30946 25.40962 25.50275 25.58980
#> 20 13.34088 13.34079 13.34072 13.34066 13.41328 13.52989 13.63819 13.73958
#> 21 23.86502 23.86500 23.86498 23.86497 23.94856 24.08147 24.20497 24.32050
#>      0.1995   0.1861   0.1735   0.1618   0.1509   0.1407   0.1313   0.1224
#> 1  24.40508 24.40546 24.40570 24.40585 23.88586 23.88580 23.88574 23.88570
#> 2  18.31203 18.31208 18.31210 18.31209 19.03677 19.03670 19.03664 19.03660
#> 3  18.24113 18.24142 18.24162 18.24175 18.19298 18.19283 18.19272 18.19263
#> 4  13.17415 13.17407 13.17404 13.17404 14.03473 14.03493 14.03508 14.03520
#> 5  22.88188 22.88172 22.88164 22.88160 23.00978 23.00954 23.00935 23.00920
#> 6  22.58674 22.58650 22.58638 22.58635 23.60195 23.60224 23.60247 23.60265
#> 7  21.36694 21.36668 21.36655 21.36651 20.93195 20.93186 20.93179 20.93173
#> 8  16.03558 16.03567 16.03576 16.03584 15.61700 15.61700 15.61700 15.61700
#> 9  16.03558 16.03567 16.03576 16.03584 15.63549 15.63553 15.63556 15.63558
#> 10 11.83200 11.83185 11.83166 11.83147 11.72908 11.72899 11.72892 11.72887
#> 11 11.90856 11.90840 11.90822 11.90805 11.23013 11.23007 11.23003 11.22999
#> 12 12.34695 12.34670 12.34649 12.34632 11.39294 11.39299 11.39303 11.39306
#> 13 28.59454 28.59430 28.59412 28.59398 28.75099 28.75112 28.75123 28.75131
#> 14 22.35319 22.35311 22.35312 22.35319 22.42860 22.42880 22.42896 22.42909
#> 15 15.47260 15.47277 15.47286 15.47290 15.90909 15.90880 15.90858 15.90841
#> 16 16.60765 16.60761 16.60757 16.60754 17.07821 17.07810 17.07801 17.07794
#> 17 14.54726 14.54690 14.54669 14.54657 14.53112 14.53139 14.53159 14.53175
#> 18 26.16817 26.16844 26.16860 26.16870 25.95863 25.95858 25.95854 25.95851
#> 19 25.61316 25.61318 25.61316 25.61313 26.17159 26.17133 26.17113 26.17098
#> 20 13.76566 13.76611 13.76646 13.76672 14.34450 14.34476 14.34497 14.34512
#> 21 24.35114 24.35136 24.35150 24.35158 23.12862 23.12864 23.12865 23.12866
#>      0.1142   0.1065   0.0993   0.0926   0.0864   0.0805   0.0751   0.0700
#> 1  23.88567 23.88564 23.88562 23.88561 23.88560 23.88559 23.88558 23.88558
#> 2  19.03656 19.03654 19.03652 19.03650 19.03649 19.03648 19.03647 19.03646
#> 3  18.19256 18.19251 18.19247 18.19244 18.19242 18.19240 18.19239 18.19237
#> 4  14.03529 14.03536 14.03542 14.03546 14.03549 14.03552 14.03554 14.03555
#> 5  23.00909 23.00901 23.00894 23.00889 23.00885 23.00882 23.00880 23.00878
#> 6  23.60279 23.60289 23.60297 23.60304 23.60308 23.60312 23.60315 23.60317
#> 7  20.93169 20.93166 20.93164 20.93162 20.93161 20.93159 20.93159 20.93158
#> 8  15.61700 15.61700 15.61700 15.61700 15.61700 15.61700 15.61700 15.61700
#> 9  15.63559 15.63561 15.63562 15.63562 15.63563 15.63563 15.63564 15.63564
#> 10 11.72883 11.72880 11.72878 11.72876 11.72874 11.72873 11.72872 11.72872
#> 11 11.22996 11.22994 11.22993 11.22992 11.22991 11.22990 11.22989 11.22989
#> 12 11.39308 11.39310 11.39311 11.39312 11.39313 11.39314 11.39314 11.39315
#> 13 28.75137 28.75142 28.75145 28.75148 28.75150 28.75152 28.75153 28.75154
#> 14 22.42918 22.42925 22.42931 22.42935 22.42939 22.42941 22.42943 22.42945
#> 15 15.90828 15.90818 15.90810 15.90804 15.90800 15.90796 15.90793 15.90791
#> 16 17.07789 17.07785 17.07782 17.07779 17.07777 17.07776 17.07775 17.07774
#> 17 14.53188 14.53197 14.53205 14.53210 14.53215 14.53218 14.53221 14.53223
#> 18 25.95849 25.95847 25.95846 25.95845 25.95844 25.95843 25.95843 25.95842
#> 19 26.17086 26.17077 26.17070 26.17065 26.17060 26.17057 26.17055 26.17053
#> 20 14.34524 14.34534 14.34541 14.34546 14.34550 14.34554 14.34556 14.34558
#> 21 23.12867 23.12868 23.12869 23.12869 23.12869 23.12870 23.12870 23.12870
#>      0.0653   0.0609   0.0568   0.0530   0.0494   0.0461   0.0430   0.0401
#> 1  24.10579 24.10557 24.10539 24.10525 24.10515 24.10507 24.10502 24.10498
#> 2  19.03717 19.03699 19.03688 19.03682 19.03680 19.03679 19.03680 19.03681
#> 3  18.26369 18.26335 18.26314 18.26302 18.26295 18.26293 18.26293 18.26294
#> 4  13.93900 13.93898 13.93898 13.93898 13.93899 13.93899 13.93899 13.93900
#> 5  22.78824 22.78841 22.78858 22.78874 22.78888 22.78899 22.78908 22.78915
#> 6  23.49137 23.49165 23.49182 23.49193 23.49198 23.49201 23.49201 23.49200
#> 7  20.65773 20.65807 20.65835 20.65857 20.65874 20.65886 20.65895 20.65901
#> 8  15.59646 15.59635 15.59629 15.59626 15.59625 15.59625 15.59625 15.59626
#> 9  15.63495 15.63484 15.63478 15.63473 15.63471 15.63470 15.63470 15.63470
#> 10 11.84385 11.84382 11.84378 11.84375 11.84371 11.84368 11.84366 11.84364
#> 11 11.34785 11.34788 11.34788 11.34785 11.34782 11.34779 11.34777 11.34774
#> 12 11.47662 11.47677 11.47684 11.47685 11.47684 11.47682 11.47678 11.47675
#> 13 28.83083 28.83110 28.83123 28.83127 28.83127 28.83124 28.83119 28.83115
#> 14 22.35421 22.35427 22.35431 22.35433 22.35433 22.35433 22.35433 22.35432
#> 15 15.85755 15.85725 15.85710 15.85704 15.85704 15.85707 15.85712 15.85717
#> 16 17.01099 17.01088 17.01084 17.01084 17.01086 17.01088 17.01091 17.01094
#> 17 14.39858 14.39886 14.39904 14.39915 14.39920 14.39923 14.39923 14.39922
#> 18 26.14051 26.14036 26.14023 26.14012 26.14004 26.13998 26.13993 26.13990
#> 19 26.07187 26.07200 26.07212 26.07223 26.07232 26.07240 26.07246 26.07250
#> 20 14.38758 14.38747 14.38739 14.38732 14.38727 14.38723 14.38719 14.38717
#> 21 23.36515 23.36511 23.36504 23.36494 23.36485 23.36477 23.36470 23.36465
#>      0.0374   0.0349   0.0325   0.0303   0.0283   0.0264   0.0246   0.0229
#> 1  23.87068 23.87063 23.87058 23.87054 23.87050 23.87046 23.87042 23.87039
#> 2  19.44626 19.44643 19.44659 19.44674 19.44689 19.44704 19.44717 19.44731
#> 3  18.27329 18.27331 18.27333 18.27335 18.27337 18.27338 18.27340 18.27342
#> 4  14.18108 14.18124 14.18139 14.18154 14.18168 14.18181 14.18194 14.18206
#> 5  22.82448 22.82435 22.82423 22.82411 22.82399 22.82389 22.82378 22.82369
#> 6  23.51032 23.51043 23.51053 23.51062 23.51071 23.51080 23.51089 23.51097
#> 7  20.64298 20.64275 20.64253 20.64232 20.64212 20.64193 20.64175 20.64157
#> 8  15.23303 15.23292 15.23282 15.23272 15.23263 15.23254 15.23245 15.23237
#> 9  15.26729 15.26720 15.26712 15.26704 15.26696 15.26688 15.26680 15.26673
#> 10 12.28623 12.28639 12.28653 12.28667 12.28680 12.28692 12.28704 12.28715
#> 11 11.41305 11.41306 11.41306 11.41307 11.41307 11.41308 11.41308 11.41308
#> 12 11.22811 11.22801 11.22791 11.22782 11.22773 11.22765 11.22757 11.22750
#> 13 29.41402 29.41418 29.41432 29.41446 29.41458 29.41470 29.41481 29.41491
#> 14 22.13192 22.13188 22.13183 22.13179 22.13175 22.13171 22.13168 22.13164
#> 15 15.98397 15.98403 15.98408 15.98413 15.98418 15.98423 15.98427 15.98431
#> 16 17.07436 17.07442 17.07448 17.07453 17.07458 17.07463 17.07468 17.07473
#> 17 14.20403 14.20397 14.20392 14.20387 14.20382 14.20378 14.20374 14.20371
#> 18 26.17185 26.17188 26.17190 26.17193 26.17195 26.17197 26.17199 26.17201
#> 19 25.84621 25.84613 25.84606 25.84599 25.84593 25.84587 25.84581 25.84576
#> 20 14.57982 14.57995 14.58008 14.58020 14.58031 14.58042 14.58052 14.58061
#> 21 23.01699 23.01685 23.01670 23.01657 23.01644 23.01632 23.01620 23.01609
#>      0.0214   0.0199   0.0186   0.0174   0.0162   0.0151   0.0141   0.0131
#> 1  23.87036 23.87033 23.87030 23.87028 23.87026 23.87024 23.87022 23.87020
#> 2  19.44743 19.44756 19.44767 19.44779 19.44790 19.44800 19.44810 19.44819
#> 3  18.27343 18.27345 18.27346 18.27348 18.27349 18.27350 18.27351 18.27353
#> 4  14.18218 14.18229 14.18240 14.18251 14.18260 14.18270 14.18279 14.18288
#> 5  22.82359 22.82350 22.82342 22.82333 22.82326 22.82318 22.82311 22.82304
#> 6  23.51104 23.51112 23.51119 23.51126 23.51132 23.51139 23.51145 23.51151
#> 7  20.64140 20.64124 20.64108 20.64093 20.64079 20.64065 20.64052 20.64040
#> 8  15.23229 15.23221 15.23213 15.23206 15.23199 15.23192 15.23186 15.23180
#> 9  15.26666 15.26660 15.26653 15.26647 15.26641 15.26635 15.26629 15.26624
#> 10 12.28725 12.28735 12.28744 12.28753 12.28761 12.28769 12.28776 12.28783
#> 11 11.41308 11.41308 11.41309 11.41309 11.41309 11.41309 11.41309 11.41309
#> 12 11.22743 11.22737 11.22731 11.22725 11.22720 11.22715 11.22710 11.22706
#> 13 29.41501 29.41510 29.41518 29.41526 29.41533 29.41540 29.41546 29.41551
#> 14 22.13161 22.13158 22.13155 22.13152 22.13150 22.13147 22.13145 22.13143
#> 15 15.98435 15.98439 15.98443 15.98446 15.98449 15.98452 15.98455 15.98458
#> 16 17.07477 17.07481 17.07485 17.07488 17.07492 17.07495 17.07498 17.07501
#> 17 14.20367 14.20364 14.20362 14.20359 14.20357 14.20356 14.20354 14.20352
#> 18 26.17203 26.17204 26.17206 26.17207 26.17209 26.17210 26.17211 26.17212
#> 19 25.84571 25.84567 25.84563 25.84559 25.84556 25.84552 25.84549 25.84546
#> 20 14.58071 14.58079 14.58087 14.58095 14.58103 14.58109 14.58116 14.58122
#> 21 23.01598 23.01588 23.01579 23.01569 23.01561 23.01553 23.01545 23.01537
#>      0.0122   0.0114   0.0106   0.0099   0.0093   0.0086   0.0081   0.0075
#> 1  23.87019 23.87018 23.87016 23.87015 23.87014 23.87013 23.87013 23.87012
#> 2  19.44828 19.44837 19.44846 19.44854 19.44861 19.44869 19.44876 19.44882
#> 3  18.27354 18.27355 18.27356 18.27357 18.27358 18.27359 18.27360 18.27361
#> 4  14.18296 14.18304 14.18311 14.18318 14.18325 14.18332 14.18338 14.18344
#> 5  22.82298 22.82292 22.82286 22.82280 22.82275 22.82270 22.82265 22.82261
#> 6  23.51156 23.51161 23.51166 23.51171 23.51176 23.51180 23.51184 23.51189
#> 7  20.64028 20.64017 20.64006 20.63996 20.63986 20.63976 20.63967 20.63959
#> 8  15.23174 15.23168 15.23163 15.23157 15.23152 15.23147 15.23143 15.23138
#> 9  15.26618 15.26613 15.26608 15.26604 15.26599 15.26595 15.26590 15.26586
#> 10 12.28790 12.28796 12.28801 12.28807 12.28812 12.28816 12.28821 12.28825
#> 11 11.41309 11.41309 11.41309 11.41309 11.41309 11.41308 11.41308 11.41308
#> 12 11.22701 11.22697 11.22694 11.22690 11.22687 11.22684 11.22681 11.22678
#> 13 29.41557 29.41562 29.41566 29.41570 29.41574 29.41578 29.41581 29.41584
#> 14 22.13141 22.13139 22.13137 22.13136 22.13134 22.13133 22.13131 22.13130
#> 15 15.98460 15.98462 15.98465 15.98467 15.98469 15.98471 15.98473 15.98474
#> 16 17.07504 17.07507 17.07510 17.07512 17.07515 17.07517 17.07519 17.07521
#> 17 14.20351 14.20350 14.20349 14.20348 14.20347 14.20347 14.20346 14.20346
#> 18 26.17213 26.17214 26.17215 26.17216 26.17217 26.17218 26.17218 26.17219
#> 19 25.84544 25.84542 25.84539 25.84537 25.84536 25.84534 25.84533 25.84531
#> 20 14.58128 14.58134 14.58139 14.58144 14.58149 14.58153 14.58157 14.58161
#> 21 23.01530 23.01524 23.01517 23.01511 23.01506 23.01500 23.01495 23.01490
#>      0.0070   0.0065   0.0061   0.0057   0.0053   0.0049
#> 1  23.87011 23.87011 23.87010 23.87010 23.87010 23.87009
#> 2  19.44889 19.44895 19.44901 19.44907 19.44912 19.44917
#> 3  18.27361 18.27362 18.27363 18.27364 18.27364 18.27365
#> 4  14.18350 14.18355 14.18360 14.18365 14.18370 14.18375
#> 5  22.82256 22.82252 22.82248 22.82245 22.82241 22.82238
#> 6  23.51192 23.51196 23.51200 23.51203 23.51206 23.51209
#> 7  20.63950 20.63943 20.63935 20.63928 20.63921 20.63915
#> 8  15.23134 15.23130 15.23126 15.23122 15.23118 15.23115
#> 9  15.26583 15.26579 15.26575 15.26572 15.26568 15.26565
#> 10 12.28829 12.28833 12.28836 12.28839 12.28842 12.28845
#> 11 11.41308 11.41308 11.41308 11.41308 11.41308 11.41307
#> 12 11.22676 11.22674 11.22671 11.22669 11.22667 11.22666
#> 13 29.41586 29.41589 29.41591 29.41593 29.41595 29.41597
#> 14 22.13129 22.13127 22.13126 22.13125 22.13124 22.13123
#> 15 15.98476 15.98477 15.98479 15.98480 15.98481 15.98482
#> 16 17.07523 17.07525 17.07526 17.07528 17.07530 17.07531
#> 17 14.20346 14.20346 14.20346 14.20346 14.20346 14.20346
#> 18 26.17220 26.17220 26.17221 26.17221 26.17221 26.17222
#> 19 25.84530 25.84529 25.84528 25.84527 25.84526 25.84526
#> 20 14.58165 14.58169 14.58172 14.58175 14.58178 14.58181
#> 21 23.01486 23.01481 23.01477 23.01473 23.01470 23.01466
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 22.8 21.4 18.1 14.3 24.4 22.8 19.2 17.3 15.2 10.4 10.4 14.7 30.4 21.5 15.5
#> [16] 15.2 13.3 27.3 26.0 15.0 21.4
#> 
#> $X
#>               am       carb        cyl       disp        drat       gear
#>  [1,]  1.5811388 -1.0471135 -1.1893017 -1.0955622  0.56403122  0.7179582
#>  [2,] -0.6324555 -1.0471135 -0.1034175  0.1149825 -0.79925501 -0.7897540
#>  [3,] -0.6324555 -1.0471135 -0.1034175 -0.1513373 -1.36581552 -0.7897540
#>  [4,] -0.6324555  0.7853351  0.9824666  0.9381530 -0.56908980 -0.7897540
#>  [5,] -0.6324555 -0.4362973 -1.1893017 -0.7832417  0.28075097  0.7179582
#>  [6,] -0.6324555 -0.4362973 -1.1893017 -0.8308564  0.68796634  0.7179582
#>  [7,] -0.6324555  0.7853351 -0.1034175 -0.6145724  0.68796634  0.7179582
#>  [8,] -0.6324555  0.1745189  0.9824666  0.2586338 -0.81696002 -0.7897540
#>  [9,] -0.6324555  0.1745189  0.9824666  0.2586338 -0.81696002 -0.7897540
#> [10,] -0.6324555  0.7853351  0.9824666  1.8420264 -1.06483025 -0.7897540
#> [11,] -0.6324555  0.7853351  0.9824666  1.7451828 -0.94089514 -0.7897540
#> [12,] -0.6324555  0.7853351  0.9824666  1.5837768 -0.53367977 -0.7897540
#> [13,]  1.5811388 -0.4362973 -1.1893017 -1.3562328  2.47617295  0.7179582
#> [14,] -0.6324555 -1.0471135 -1.1893017 -0.9979116  0.29845598 -0.7897540
#> [15,] -0.6324555 -0.4362973  0.9824666  0.5992004 -1.36581552 -0.7897540
#> [16,] -0.6324555 -0.4362973  0.9824666  0.4862163 -0.67531990 -0.7897540
#> [17,] -0.6324555  0.7853351  0.9824666  0.8574500  0.35157103 -0.7897540
#> [18,]  1.5811388 -1.0471135 -1.1893017 -1.3296009  0.97124659  0.7179582
#> [19,]  1.5811388 -0.4362973 -1.1893017 -0.9962975  1.59092215  2.2256703
#> [20,]  1.5811388  3.2286000  0.9824666  0.4620054  0.01517573  2.2256703
#> [21,]  1.5811388 -0.4362973 -1.1893017 -0.9906483  1.02436164  0.7179582
#>                 hp        qsec         vs          wt
#>  [1,] -0.777129550  0.25962375  1.0488088 -1.06140517
#>  [2,] -0.543006977  0.72280101  1.0488088 -0.17296168
#>  [3,] -0.611866557  1.15807604  1.0488088  0.07024352
#>  [4,]  1.316201693 -1.28616063 -0.9534626  0.17943770
#>  [5,] -1.204058948  1.03530616  1.0488088 -0.19777854
#>  [6,] -0.749585718  2.65363637  1.0488088 -0.23748551
#>  [7,] -0.363972068  0.08662983  1.0488088  0.05039004
#>  [8,]  0.421027149 -0.30400160 -0.9534626  0.33826558
#>  [9,]  0.421027149 -0.08078364 -0.9534626  0.38789930
#> [10,]  0.765325050 -0.09194454 -0.9534626  1.84713052
#> [11,]  0.903044211 -0.18123172 -0.9534626  2.01985584
#> [12,]  1.109622952 -0.40444969 -0.9534626  1.94143457
#> [13,] -1.341778109  0.20939971  1.0488088 -1.76124055
#> [14,] -0.722041886  1.04088661  1.0488088 -0.91746740
#> [15,]  0.007869666 -0.71137438 -0.9534626  0.12980398
#> [16,]  0.007869666 -0.47141507 -0.9534626  0.04542667
#> [17,]  1.316201693 -1.52611994 -0.9534626  0.44745976
#> [18,] -1.148971284  0.42145677  1.0488088 -1.44358478
#> [19,] -0.804673382 -0.80624201 -0.9534626 -1.24008655
#> [20,]  2.555674140 -1.97813630 -0.9534626  0.17943770
#> [21,] -0.556778893  0.25404330  1.0488088 -0.60477500
#> attr(,"center")
#>  [1]   0.2857143   2.7142857   6.1904762 243.7523810   3.5314286   3.5238095
#>  [7] 149.4285714  18.1447619   0.4761905   3.3892381
#> attr(,"scale")
#>  [1]   0.4517540   1.6371537   1.8418171 123.9111572   0.5648117   0.6632566
#>  [7]  72.6115375   1.7919705   0.4994328   1.0073798
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 20
#> 
#> $lambda.min
#> [1] 1.312565
#> 
#> $null.dev
#> [1] 28.82122
#> 
#> $Bias
#> [1] 1.634133
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 21.87445