Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 25.929458 24.724870 22.871888 20.420121 18.378212 16.702410 15.338281
#>   [8] 14.238935 13.364011 12.678805 12.153500 11.762507 11.483890 11.298871
#>  [15] 11.191397 11.137789 10.912057 10.643918 10.526268 10.558520 10.615389
#>  [22] 10.650683 10.226442 10.094938  9.756319  9.229642  8.460127  7.480383
#>  [29]  7.434143  7.701419  7.763062  7.822610  7.855069  6.574269  5.020459
#>  [36]  5.031909  5.048556  5.065809  4.778727  4.802021  4.863201  5.002633
#>  [43]  5.184104  5.146543  5.120963  5.204175  5.108248  5.240626  5.240869
#>  [50]  5.241048  5.241185  5.241290  6.196185  6.196605  6.196953  6.249054
#>  [57]  7.662310  8.594680  8.595419  8.703812  8.693603  8.732517  8.766717
#>  [64]  8.773103  8.773685  8.876932 10.266829 10.820757 11.056708 11.097969
#>  [71] 12.618760 12.615741 12.839879 12.841466 12.828201 12.803728 13.367742
#>  [78] 13.368903 13.370252 13.376738 13.377569 14.998031 15.020491 15.040021
#>  [85] 15.041476 15.042915 15.044338 15.179347 15.181261 15.182859 15.184431
#>  [92] 15.188710 15.820035 15.807560 15.814629 15.816559 15.818468 15.820395
#>  [99] 15.822291 15.834650
#> 
#> $cvse
#>   [1] 7.668527 7.663589 7.269016 6.538485 5.898148 5.338738 4.848261 4.417857
#>   [9] 4.041277 3.714381 3.434647 3.200654 3.011537 2.866417 2.763866 2.700764
#>  [17] 2.629249 2.535389 2.506279 2.504827 2.503502 2.497503 2.414610 2.381444
#>  [25] 2.386983 2.407343 2.367820 2.360150 2.348997 2.370747 2.371856 2.373843
#>  [33] 2.366996 2.281014 1.633441 1.626251 1.619528 1.613611 1.102585 1.107619
#>  [41] 1.110754 1.107046 1.105744 1.127541 1.139983 1.144999 1.173306 1.166434
#>  [49] 1.166497 1.166544 1.166584 1.166619 1.676300 1.676514 1.676697 1.687628
#>  [57] 1.994895 2.508077 2.508549 2.511555 2.511799 2.527889 2.525347 2.524172
#>  [65] 2.524498 2.549223 3.432272 3.483423 3.462704 3.456757 3.758834 3.759988
#>  [73] 3.737763 3.738300 3.738864 3.731512 3.939145 3.939684 3.940285 3.959876
#>  [81] 3.960398 4.620446 4.620098 4.620209 4.620908 4.621593 4.622263 4.607953
#>  [89] 4.611304 4.611917 4.612519 4.612744 4.968363 4.955497 4.955527 4.956215
#>  [97] 4.956898 4.957572 4.958243 4.957581
#> 
#> $fold
#>  [1]  4  9  5 10  2  5  1  6  8 10  7  1  6  2  3  7  1  8  4  9  3
#> 
#> $lambda
#>   [1] 4.570350288 4.262323975 3.975057605 3.707152026 3.457302387 3.224291777
#>   [7] 3.006985303 2.804324558 2.615322468 2.439058487 2.274674108 2.121368686
#>  [13] 1.978395537 1.845058299 1.720707545 1.604737615 1.496583671 1.395718941
#>  [19] 1.301652155 1.213925156 1.132110664 1.055810194 0.984652121 0.918289863
#>  [25] 0.856400200 0.798681693 0.744853220 0.694652606 0.647835345 0.604173411
#>  [31] 0.563454145 0.525479222 0.490063681 0.457035029 0.426232398 0.397505760
#>  [37] 0.370715203 0.345730239 0.322429178 0.300698530 0.280432456 0.261532247
#>  [43] 0.243905849 0.227467411 0.212136869 0.197839554 0.184505830 0.172070755
#>  [49] 0.160473762 0.149658367 0.139571894 0.130165216 0.121392516 0.113211066
#>  [55] 0.105581019 0.098465212 0.091828986 0.085640020 0.079868169 0.074485322
#>  [61] 0.069465261 0.064783534 0.060417341 0.056345415 0.052547922 0.049006368
#>  [67] 0.045703503 0.042623240 0.039750576 0.037071520 0.034573024 0.032242918
#>  [73] 0.030069853 0.028043246 0.026153225 0.024390585 0.022746741 0.021213687
#>  [79] 0.019783955 0.018450583 0.017207075 0.016047376 0.014965837 0.013957189
#>  [85] 0.013016522 0.012139252 0.011321107 0.010558102 0.009846521 0.009182899
#>  [91] 0.008564002 0.007986817 0.007448532 0.006946526 0.006478353 0.006041734
#>  [97] 0.005634541 0.005254792 0.004900637 0.004570350
#> 
#> $fit
#> $beta
#>               4.5704     4.2623     3.9751   3.7072    3.4573    3.2243
#> (Intercept) 19.23333 20.7811595 22.2246673 23.57089 24.826377 25.997251
#> am           0.00000  0.0000000  0.0000000  0.00000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.00000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.0000000  0.00000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.0000000  0.00000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.00000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.00000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.00000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.00000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.00000  0.000000  0.000000
#> wt           0.00000 -0.4575693 -0.8842999 -1.28227 -1.653419 -1.999553
#>                3.0070    2.8043   2.6153    2.4391    2.2747    2.1214
#> (Intercept) 27.089211 28.107578 29.05731 29.943033 30.769061 31.539417
#> am           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> wt          -2.322359 -2.623409 -2.90417 -3.166008 -3.410199 -3.637932
#>                1.9784    1.8451    1.7207    1.6047    1.4966    1.3957
#> (Intercept) 32.257855 32.927872 33.552732 34.135478 34.543947 34.543947
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -3.850317 -4.048388 -4.233109 -4.405381 -4.526133 -4.526133
#>                1.3017    1.2139    1.1321    1.0558      0.9847     0.9183
#> (Intercept) 34.543947 34.543947 34.543947 34.543947 33.49238571 32.4102506
#> am           0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.05740334  0.1164757
#> vs           0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> wt          -4.526133 -4.526133 -4.526133 -4.526133 -4.51933256 -4.5123349
#>                 0.8564    0.7987    0.7449    0.6947    0.6478    0.6042
#> (Intercept) 35.8226922 36.433522 36.836991 36.837281 36.837447 36.837541
#> am           0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -0.7161648 -1.058262 -1.284225 -1.284388 -1.284481 -1.284534
#> disp         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -3.5733883 -3.118282 -2.817673 -2.817457 -2.817334 -2.817263
#>                0.5635    0.5255    0.4901    0.4570    0.4262    0.3975
#> (Intercept) 36.837596 36.837627 36.837644 36.837654 36.837660 36.837664
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.284564 -1.284581 -1.284591 -1.284597 -1.284600 -1.284602
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.817223 -2.817200 -2.817186 -2.817179 -2.817174 -2.817172
#>                0.3707    0.3457       0.3224      0.3007      0.2804
#> (Intercept) 36.837665 36.837667 36.838183073 36.84278701 36.84733640
#> am           0.000000  0.000000  0.000000000  0.00000000  0.00000000
#> carb         0.000000  0.000000 -0.004150898 -0.03263958 -0.05900923
#> cyl         -1.284603 -1.284604 -1.282979827 -1.27242692 -1.26282019
#> disp         0.000000  0.000000  0.000000000  0.00000000  0.00000000
#> drat         0.000000  0.000000  0.000000000  0.00000000  0.00000000
#> gear         0.000000  0.000000  0.000000000  0.00000000  0.00000000
#> hp           0.000000  0.000000  0.000000000  0.00000000  0.00000000
#> qsec         0.000000  0.000000  0.000000000  0.00000000  0.00000000
#> vs           0.000000  0.000000  0.000000000  0.00000000  0.00000000
#> wt          -2.817171 -2.817170 -2.816717368 -2.81282318 -2.80900414
#>                  0.2615     0.2439     0.2275     0.2121     0.1978    0.1845
#> (Intercept) 36.85159066 36.8555587 36.8592594 36.8627106 36.8659292 15.427798
#> am           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> carb        -0.08359316 -0.1065198 -0.1279013 -0.1478417 -0.1664383 -1.435237
#> cyl         -1.25387127 -1.2455259 -1.2377431 -1.2304848 -1.2237156  0.000000
#> disp         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> drat         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> gear         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  3.501237
#> hp           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> qsec         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> vs           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  1.566928
#> wt          -2.80543407 -2.8021042 -2.7989988 -2.7961027 -2.7934017 -1.593954
#>                0.1721    0.1605    0.1497    0.1396    0.1302    0.1214
#> (Intercept) 15.425606 15.423791 15.422287 15.421041 15.420010 15.419156
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb        -1.435431 -1.435591 -1.435724 -1.435834 -1.435925 -1.436001
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         3.501702  3.502088  3.502407  3.502671  3.502890  3.503071
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           1.566855  1.566794  1.566744  1.566702  1.566668  1.566639
#> wt          -1.593633 -1.593367 -1.593147 -1.592965 -1.592814 -1.592689
#>                0.1132    0.1056    0.0985    0.0918    0.0856    0.0799
#> (Intercept) 15.418448 15.417862 15.417377 15.416975 15.416642 15.416366
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb        -1.436063 -1.436115 -1.436158 -1.436194 -1.436223 -1.436247
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         3.503222  3.503346  3.503449  3.503534  3.503605  3.503664
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           1.566616  1.566596  1.566580  1.566566  1.566555  1.566546
#> wt          -1.592585 -1.592499 -1.592428 -1.592370 -1.592321 -1.592280
#>                0.0745    0.0695    0.0648       0.0604       0.0563
#> (Intercept) 15.416138 15.415949 15.415792 15.603683050 15.602330174
#> am           0.000000  0.000000  0.000000  0.000000000  0.000000000
#> carb        -1.436267 -1.436284 -1.436298 -1.453916459 -1.454036567
#> cyl          0.000000  0.000000  0.000000  0.000000000  0.000000000
#> disp         0.000000  0.000000  0.000000 -0.004487514 -0.004489765
#> drat         0.000000  0.000000  0.000000  0.000000000  0.000000000
#> gear         3.503712  3.503752  3.503785  3.431990477  3.432249956
#> hp           0.000000  0.000000  0.000000  0.000000000  0.000000000
#> qsec         0.000000  0.000000  0.000000  0.000000000  0.000000000
#> vs           1.566539  1.566532  1.566527  1.266294035  1.266133944
#> wt          -1.592247 -1.592219 -1.592196 -1.202598928 -1.202197818
#>                   0.0525       0.0490       0.0457       0.0426      0.0398
#> (Intercept) 15.601139071 15.600090721 15.599168275 15.598356815 17.41258018
#> am           0.000000000  0.000000000  0.000000000  0.000000000  0.00000000
#> carb        -1.454142328 -1.454235425 -1.454317352 -1.454389428 -1.51684175
#> cyl          0.000000000  0.000000000  0.000000000  0.000000000  0.00000000
#> disp        -0.004491738 -0.004493468 -0.004494984 -0.004496313 -0.00586998
#> drat         0.000000000  0.000000000  0.000000000  0.000000000  0.00000000
#> gear         3.432478529  3.432679805  3.432856984  3.433012907  3.42489738
#> hp           0.000000000  0.000000000  0.000000000  0.000000000  0.00000000
#> qsec         0.000000000  0.000000000  0.000000000  0.000000000 -0.11582942
#> vs           1.265993406  1.265870036  1.265761740  1.265666677  1.44970691
#> wt          -1.201845467 -1.201535975 -1.201264150 -1.201025424 -0.98384267
#>                   0.0371       0.0346       0.0322       0.0301       0.0280
#> (Intercept) 17.413643377 17.414634525 17.415558503 17.416419863 17.417222850
#> am           0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> carb        -1.516957300 -1.517065020 -1.517165442 -1.517259060 -1.517346336
#> cyl          0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> disp        -0.005872904 -0.005875629 -0.005878169 -0.005880537 -0.005882744
#> drat         0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> gear         3.424998905  3.425093559  3.425181803  3.425264069  3.425340762
#> hp           0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> qsec        -0.115935963 -0.116035285 -0.116127876 -0.116214194 -0.116294662
#> vs           1.449770281  1.449829356  1.449884431  1.449935774  1.449983639
#> wt          -0.983401746 -0.982990704 -0.982607517 -0.982250297 -0.981917285
#>                   0.0262       0.0244      0.0227       0.0212       0.0198
#> (Intercept) 17.417971420 17.418669260 17.41931981 17.419926273 17.420491637
#> am           0.000000000  0.000000000  0.00000000  0.000000000  0.000000000
#> carb        -1.517427698 -1.517503546 -1.51757426 -1.517640173 -1.517701624
#> cyl          0.000000000  0.000000000  0.00000000  0.000000000  0.000000000
#> disp        -0.005884802 -0.005886721 -0.00588851 -0.005890177 -0.005891731
#> drat         0.000000000  0.000000000  0.00000000  0.000000000  0.000000000
#> gear         3.425412260  3.425478913  3.42554105  3.425598978  3.425652981
#> hp           0.000000000  0.000000000  0.00000000  0.000000000  0.000000000
#> qsec        -0.116369678 -0.116439610 -0.11650480 -0.116565579 -0.116622236
#> vs           1.450028262  1.450069861  1.45010864  1.450144795  1.450178498
#> wt          -0.981606839 -0.981317431 -0.98104764 -0.980796123 -0.980561655
#>                  0.0185       0.0172      0.0160       0.0150       0.0140
#> (Intercept) 17.42101869 17.421510022 17.42196806 17.422395058 17.422793120
#> am           0.00000000  0.000000000  0.00000000  0.000000000  0.000000000
#> carb        -1.51775891 -1.517812316 -1.51786210 -1.517908514 -1.517951781
#> cyl          0.00000000  0.000000000  0.00000000  0.000000000  0.000000000
#> disp        -0.00589318 -0.005894531 -0.00589579 -0.005896964 -0.005898058
#> drat         0.00000000  0.000000000  0.00000000  0.000000000  0.000000000
#> gear         3.42570332  3.425750256  3.42579401  3.425834794  3.425872817
#> hp           0.00000000  0.000000000  0.00000000  0.000000000  0.000000000
#> qsec        -0.11667505 -0.116724292 -0.11677019 -0.116812985 -0.116852876
#> vs           1.45020992  1.450239208  1.45026651  1.450291969  1.450315700
#> wt          -0.98034308 -0.980139309 -0.97994935 -0.979772265 -0.979607180
#>                   0.0130       0.0121     0.0113       0.0106       0.0098
#> (Intercept) 17.423164206 17.586965507 17.5866054 17.586220944 17.585833588
#> am           0.000000000 -0.118599559 -0.1190830 -0.119442937 -0.119709128
#> carb        -1.517992116 -1.521924313 -1.5220690 -1.522183431 -1.522272633
#> cyl          0.000000000  0.000000000  0.0000000  0.000000000  0.000000000
#> disp        -0.005899079 -0.006046011 -0.0060437 -0.006041614 -0.006039746
#> drat         0.000000000  0.000000000  0.0000000  0.000000000  0.000000000
#> gear         3.425908263  3.468212893  3.4686753  3.469040891  3.469327752
#> hp           0.000000000  0.000000000  0.0000000  0.000000000  0.000000000
#> qsec        -0.116890065 -0.130251296 -0.1303117 -0.130354115 -0.130382477
#> vs           1.450337822  1.448091086  1.4481573  1.448214097  1.448261772
#> wt          -0.979453283 -0.977170010 -0.9772421 -0.977319043 -0.977398934
#>                   0.0092       0.0086       0.0080       0.0074       0.0069
#> (Intercept) 17.585455873 17.585094823 17.584753973 17.584434712 17.584137147
#> am          -0.119904111 -0.120045041 -0.120145000 -0.120213971 -0.120259586
#> carb        -1.522340985 -1.522392281 -1.522429754 -1.522456124 -1.522473660
#> cyl          0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> disp        -0.006038082 -0.006036603 -0.006035291 -0.006034128 -0.006033094
#> drat         0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> gear         3.469550960  3.469722964  3.469853985  3.469952371  3.470024905
#> hp           0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> qsec        -0.130399961 -0.130409065 -0.130411769 -0.130409643 -0.130403932
#> vs           1.448301008  1.448332661  1.448357667  1.448376957  1.448391417
#> wt          -0.977480143 -0.977561364 -0.977641567 -0.977719963 -0.977795963
#>                   0.0065       0.0060       0.0056       0.0053       0.0049
#> (Intercept) 17.583860642 17.583604165 17.583366494 17.583146344 17.582942439
#> am          -0.120287683 -0.120302724 -0.120308122 -0.120306483 -0.120299793
#> carb        -1.522484239 -1.522489396 -1.522490385 -1.522488216 -1.522483701
#> cyl          0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> disp        -0.006032175 -0.006031356 -0.006030624 -0.006029968 -0.006029379
#> drat         0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> gear         3.470077075  3.470113301  3.470137127  3.470151388  3.470158337
#> hp           0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> qsec        -0.130395616 -0.130385469 -0.130374096 -0.130361966 -0.130349442
#> vs           1.448401856  1.448408996  1.448413465  1.448415801  1.448416460
#> wt          -0.977869145 -0.977939224 -0.978006024 -0.978069458 -0.978129501
#>                   0.0046
#> (Intercept) 17.582753555
#> am          -0.120289561
#> carb        -1.522477489
#> cyl          0.000000000
#> disp        -0.006028848
#> drat         0.000000000
#> gear         3.470159762
#> hp           0.000000000
#> qsec        -0.130336800
#> vs           1.448415820
#> wt          -0.978186186
#> 
#> $iter
#>   [1]   0   4   3   3   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   4   4  51  35  14   1   1   1   1   1   1   1   1   1
#>  [37]   1   1   4   5   5   5   5   5   5   5 115   1   1   1   1   1   1   1
#>  [55]   1   1   1   1   1   1   1   1  48   1   1   1   1   1  67   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1  24   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.570350288 4.262323975 3.975057605 3.707152026 3.457302387 3.224291777
#>   [7] 3.006985303 2.804324558 2.615322468 2.439058487 2.274674108 2.121368686
#>  [13] 1.978395537 1.845058299 1.720707545 1.604737615 1.496583671 1.395718941
#>  [19] 1.301652155 1.213925156 1.132110664 1.055810194 0.984652121 0.918289863
#>  [25] 0.856400200 0.798681693 0.744853220 0.694652606 0.647835345 0.604173411
#>  [31] 0.563454145 0.525479222 0.490063681 0.457035029 0.426232398 0.397505760
#>  [37] 0.370715203 0.345730239 0.322429178 0.300698530 0.280432456 0.261532247
#>  [43] 0.243905849 0.227467411 0.212136869 0.197839554 0.184505830 0.172070755
#>  [49] 0.160473762 0.149658367 0.139571894 0.130165216 0.121392516 0.113211066
#>  [55] 0.105581019 0.098465212 0.091828986 0.085640020 0.079868169 0.074485322
#>  [61] 0.069465261 0.064783534 0.060417341 0.056345415 0.052547922 0.049006368
#>  [67] 0.045703503 0.042623240 0.039750576 0.037071520 0.034573024 0.032242918
#>  [73] 0.030069853 0.028043246 0.026153225 0.024390585 0.022746741 0.021213687
#>  [79] 0.019783955 0.018450583 0.017207075 0.016047376 0.014965837 0.013957189
#>  [85] 0.013016522 0.012139252 0.011321107 0.010558102 0.009846521 0.009182899
#>  [91] 0.008564002 0.007986817 0.007448532 0.006946526 0.006478353 0.006041734
#>  [97] 0.005634541 0.005254792 0.004900637 0.004570350
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 24
#> 
#> $loss
#>   [1] 545.88667 461.67910 391.22696 332.55071 283.94136 243.92428 211.22795
#>   [8] 184.75674 163.56728 146.84803 133.90148 124.12872 117.01612 112.12368
#>  [15] 109.07505 107.54874 107.23653 107.23653 107.23653 107.23653 107.23653
#>  [22] 107.23653 103.14067  99.34014  69.56063  61.84160  60.38716  60.38715
#>  [29]  60.38715  60.38715  60.38715  60.38715  60.38715  60.38715  60.38715
#>  [36]  60.38715  60.38715  60.38715  60.29283  59.68887  59.19738  58.79765
#>  [43]  58.47573  58.21975  58.01949  57.86618  42.49081  42.49080  42.49080
#>  [50]  42.49079  42.49079  42.49079  42.49079  42.49079  42.49079  42.49079
#>  [57]  42.49079  42.49079  42.49079  42.49079  42.49079  42.49079  41.78451
#>  [64]  41.78450  41.78450  41.78449  41.78449  41.78449  41.61285  41.61284
#>  [71]  41.61284  41.61283  41.61282  41.61282  41.61282  41.61281  41.61281
#>  [78]  41.61281  41.61280  41.61280  41.61280  41.61280  41.61280  41.61280
#>  [85]  41.61279  41.59753  41.59752  41.59751  41.59751  41.59751  41.59751
#>  [92]  41.59751  41.59751  41.59751  41.59751  41.59751  41.59750  41.59750
#>  [99]  41.59750  41.59750
#> 
#> $linear.predictors
#>      4.5704   4.2623   3.9751   3.7072   3.4573   3.2243   3.0070   2.8043
#> 1  19.23333 19.46565 19.68231 19.88436 20.07280 20.24854 20.41243 20.56528
#> 2  19.23333 19.31007 19.38164 19.44839 19.51064 19.56869 19.62283 19.67332
#> 3  19.23333 19.20712 19.18268 19.15988 19.13862 19.11879 19.10030 19.08305
#> 4  19.23333 19.32151 19.40375 19.48045 19.55197 19.61868 19.68089 19.73890
#> 5  19.23333 19.33982 19.43912 19.53174 19.61811 19.69866 19.77378 19.84384
#> 6  19.23333 19.20712 19.18268 19.15988 19.13862 19.11879 19.10030 19.08305
#> 7  19.23333 18.91885 18.62557 18.35205 18.09696 17.85907 17.63721 17.43030
#> 8  19.23333 19.07443 18.92623 18.78802 18.65912 18.53892 18.42681 18.32226
#> 9  19.23333 19.05155 18.88201 18.72391 18.57645 18.43894 18.31069 18.19109
#> 10 19.23333 18.37892 17.58209 16.83897 16.14593 15.49960 14.89682 14.33468
#> 11 19.23333 18.29930 17.42822 16.61585 15.85823 15.15167 14.49273 13.87821
#> 12 19.23333 18.33545 17.49808 16.71715 15.98885 15.30964 14.67620 14.08545
#> 13 19.23333 19.65325 20.04487 20.41009 20.75070 21.06835 21.36460 21.64087
#> 14 19.23333 19.17052 19.11193 19.05730 19.00634 18.95882 18.91451 18.87318
#> 15 19.23333 19.20941 19.18710 19.16629 19.14688 19.12879 19.11191 19.09617
#> 16 19.23333 19.89576 20.51355 21.08969 21.62701 22.12812 22.59545 23.03128
#> 17 19.23333 19.80196 20.33227 20.82683 21.28806 21.71821 22.11936 22.49348
#> 18 19.23333 20.08886 20.88672 21.63081 22.32475 22.97193 23.57548 24.13836
#> 19 19.23333 19.51369 19.77516 20.01900 20.24641 20.45849 20.65628 20.84073
#> 20 19.23333 19.14764 19.06772 18.99318 18.92367 18.85885 18.79839 18.74201
#> 21 19.23333 19.50912 19.76631 20.00618 20.22987 20.43849 20.63305 20.81450
#>      2.6153   2.4391   2.2747   2.1214   1.9784   1.8451   1.7207   1.6047
#> 1  20.70782 20.84076 20.96474 21.08036 21.18819 21.28876 21.38254 21.47001
#> 2  19.72040 19.76432 19.80527 19.84347 19.87909 19.91231 19.94329 19.97218
#> 3  19.06697 19.05197 19.03798 19.02493 19.01277 19.00142 18.99084 18.98097
#> 4  19.79301 19.84347 19.89053 19.93441 19.97534 20.01352 20.04911 20.08231
#> 5  19.90918 19.97011 20.02694 20.07993 20.12936 20.17545 20.21844 20.25853
#> 6  19.06697 19.05197 19.03798 19.02493 19.01277 19.00142 18.99084 18.98097
#> 7  17.23734 17.05738 16.88955 16.73303 16.58707 16.45093 16.32398 16.20558
#> 8  18.22476 18.13382 18.04902 17.96993 17.89617 17.82739 17.76324 17.70341
#> 9  18.07955 17.97552 17.87851 17.78803 17.70366 17.62497 17.55158 17.48314
#> 10 13.81042 13.32149 12.86552 12.44028 12.04369 11.67384 11.32891 11.00723
#> 11 13.30509 12.77061 12.27214 11.80727 11.37374 10.96942 10.59235 10.24069
#> 12 13.53452 13.02072 12.54155 12.09467 11.67791 11.28924 10.92676 10.58872
#> 13 21.89853 22.13882 22.36292 22.57192 22.76682 22.94860 23.11812 23.27621
#> 14 18.83463 18.79869 18.76516 18.73390 18.70474 18.67755 18.65219 18.62854
#> 15 19.08149 19.06780 19.05503 19.04312 19.03202 19.02166 19.01200 19.00299
#> 16 23.43774 23.81681 24.17033 24.50002 24.80749 25.09424 25.36167 25.61107
#> 17 22.84239 23.16778 23.47124 23.75424 24.01818 24.26432 24.49388 24.70796
#> 18 24.66330 25.15286 25.60943 26.03523 26.43233 26.80266 27.14804 27.47014
#> 19 21.01276 21.17319 21.32281 21.46235 21.59248 21.71384 21.82702 21.93257
#> 20 18.68942 18.64039 18.59465 18.55200 18.51222 18.47513 18.44053 18.40827
#> 21 20.98372 21.14153 21.28871 21.42597 21.55397 21.67335 21.78469 21.88852
#>       1.4966    1.3957    1.3017    1.2139    1.1321    1.0558   0.9847
#> 1  21.531315 21.531315 21.531315 21.531315 21.531315 21.531315 21.47631
#> 2  19.992430 19.992430 19.992430 19.992430 19.992430 19.992430 20.07865
#> 3  18.974051 18.974051 18.974051 18.974051 18.974051 18.974051 18.92289
#> 4  20.105584 20.105584 20.105584 20.105584 20.105584 20.105584 20.22378
#> 5  20.286629 20.286629 20.286629 20.286629 20.286629 20.286629 20.57102
#> 6  18.974051 18.974051 18.974051 18.974051 18.974051 18.974051 18.99636
#> 7  16.122587 16.122587 16.122587 16.122587 16.122587 16.122587 16.09752
#> 8  17.661472 17.661472 17.661472 17.661472 17.661472 17.661472 17.64557
#> 9  17.435166 17.435166 17.435166 17.435166 17.435166 17.435166 17.44257
#> 10 10.781751 10.781751 10.781751 10.781751 10.781751 10.781751 10.79800
#> 11  9.994204  9.994204  9.994204  9.994204  9.994204  9.994204 10.00245
#> 12 10.351768 10.351768 10.351768 10.351768 10.351768 10.351768 10.33652
#> 13 23.387030 23.387030 23.387030 23.387030 23.387030 23.387030 23.50087
#> 14 18.611960 18.611960 18.611960 18.611960 18.611960 18.611960 18.55273
#> 15 18.996681 18.996681 18.996681 18.996681 18.996681 18.996681 18.96156
#> 16 25.785880 25.785880 25.785880 25.785880 25.785880 25.785880 25.83240
#> 17 24.858023 24.858023 24.858023 24.858023 24.858023 24.858023 24.77965
#> 18 27.695908 27.695908 27.695908 27.695908 27.695908 27.695908 27.62475
#> 19 22.006559 22.006559 22.006559 22.006559 22.006559 22.006559 21.86359
#> 20 18.385653 18.385653 18.385653 18.385653 18.385653 18.385653 18.19646
#> 21 21.961298 21.961298 21.961298 21.961298 21.961298 21.961298 21.99634
#>      0.9183   0.8564   0.7987   0.7449   0.6947   0.6478   0.6042   0.5635
#> 1  21.41970 21.25221 21.11889 21.03083 21.03076 21.03073 21.03071 21.03070
#> 2  20.16738 20.03726 20.05867 20.07282 20.07283 20.07283 20.07284 20.07284
#> 3  18.87023 17.80092 17.24054 16.87039 16.87013 16.86997 16.86989 16.86984
#> 4  20.34542 21.55892 22.25316 22.71171 22.71204 22.71223 22.71234 22.71240
#> 5  20.86369 21.70186 22.37789 22.82442 22.82474 22.82492 22.82503 22.82509
#> 6  19.01932 19.23325 19.35706 19.43884 19.43890 19.43893 19.43895 19.43897
#> 7  16.07172 15.54968 15.27602 15.09526 15.09513 15.09505 15.09501 15.09499
#> 8  17.62921 16.76464 16.33624 16.05327 16.05306 16.05295 16.05288 16.05284
#> 9  17.45019 16.58597 16.18032 15.91238 15.91219 15.91208 15.91202 15.91198
#> 10 10.81473 11.33309 11.59645 11.77040 11.77053 11.77060 11.77064 11.77066
#> 11 10.01094 10.71132 11.05387 11.28013 11.28029 11.28038 11.28044 11.28047
#> 12 10.32083 10.99361 11.30021 11.50272 11.50287 11.50295 11.50300 11.50303
#> 13 23.61802 24.14963 24.51391 24.75452 24.75470 24.75480 24.75485 24.75489
#> 14 18.49178 17.51505 16.99107 16.64498 16.64473 16.64459 16.64451 16.64446
#> 15 18.92541 17.81879 17.25613 16.88448 16.88421 16.88406 16.88397 16.88392
#> 16 25.88027 26.04353 26.16660 26.24789 26.24795 26.24798 26.24800 26.24801
#> 17 24.69900 25.31098 25.52735 25.67027 25.67037 25.67043 25.67046 25.67048
#> 18 27.55153 27.55150 27.48251 27.43695 27.43692 27.43690 27.43689 27.43688
#> 19 21.71646 21.62742 21.44631 21.32668 21.32660 21.32655 21.32652 21.32650
#> 20 18.00176 17.33638 16.83516 16.50409 16.50386 16.50372 16.50364 16.50360
#> 21 22.03241 23.02401 23.53165 23.86696 23.86720 23.86734 23.86742 23.86746
#>      0.5255   0.4901   0.4570   0.4262   0.3975   0.3707   0.3457   0.3224
#> 1  21.03069 21.03069 21.03068 21.03068 21.03068 21.03068 21.03068 21.02564
#> 2  20.07284 20.07284 20.07284 20.07284 20.07284 20.07284 20.07284 20.08041
#> 3  16.86981 16.86979 16.86978 16.86978 16.86977 16.86977 16.86977 16.87653
#> 4  22.71243 22.71245 22.71247 22.71247 22.71248 22.71248 22.71248 22.71263
#> 5  22.82512 22.82514 22.82515 22.82516 22.82516 22.82517 22.82517 22.82530
#> 6  19.43897 19.43897 19.43898 19.43898 19.43898 19.43898 19.43898 19.43419
#> 7  15.09497 15.09496 15.09496 15.09496 15.09496 15.09496 15.09495 15.09785
#> 8  16.05282 16.05281 16.05280 16.05280 16.05279 16.05279 16.05279 16.05554
#> 9  15.91196 15.91195 15.91194 15.91194 15.91194 15.91193 15.91193 15.91470
#> 10 11.77068 11.77068 11.77069 11.77069 11.77069 11.77069 11.77069 11.76997
#> 11 11.28048 11.28049 11.28050 11.28050 11.28051 11.28051 11.28051 11.27987
#> 12 11.50304 11.50305 11.50306 11.50306 11.50306 11.50306 11.50306 11.50239
#> 13 24.75490 24.75491 24.75492 24.75492 24.75493 24.75493 24.75493 24.75890
#> 14 16.64443 16.64442 16.64441 16.64440 16.64440 16.64440 16.64440 16.65120
#> 15 16.88389 16.88388 16.88387 16.88386 16.88386 16.88386 16.88386 16.89062
#> 16 26.24802 26.24802 26.24803 26.24803 26.24803 26.24803 26.24803 26.25176
#> 17 25.67049 25.67050 25.67050 25.67051 25.67051 25.67051 25.67051 25.67019
#> 18 27.43688 27.43688 27.43687 27.43687 27.43687 27.43687 27.43687 27.43627
#> 19 21.32650 21.32649 21.32649 21.32649 21.32648 21.32648 21.32648 21.31309
#> 20 16.50357 16.50356 16.50355 16.50354 16.50354 16.50354 16.50354 16.48546
#> 21 23.86749 23.86750 23.86751 23.86751 23.86752 23.86752 23.86752 23.86749
#>      0.3007   0.2804   0.2615   0.2439   0.2275   0.2121   0.1978   0.1845
#> 1  20.99080 20.95849 20.92837 20.90027 20.87407 20.84964 20.82685 19.10918
#> 2  20.13236 20.18046 20.22530 20.26712 20.30612 20.34249 20.37641 20.93864
#> 3  16.92198 16.96378 17.00274 17.03907 17.07296 17.10456 17.13403 17.57783
#> 4  22.71489 22.71731 22.71958 22.72170 22.72368 22.72552 22.72724 23.04449
#> 5  22.82741 22.82967 22.83180 22.83379 22.83564 22.83736 22.83897 23.10824
#> 6  19.40156 19.37140 19.34330 19.31709 19.29264 19.26984 19.24858 19.77552
#> 7  15.11726 15.13510 15.15172 15.16723 15.18169 15.19517 15.20774 15.13840
#> 8  16.07362 16.09016 16.10557 16.11994 16.13335 16.14584 16.15750 15.68035
#> 9  15.93298 15.94971 15.96530 15.97984 15.99340 16.00604 16.01783 15.60065
#> 10 11.76549 11.76147 11.75772 11.75422 11.75097 11.74793 11.74509 11.82230
#> 11 11.27606 11.27270 11.26957 11.26666 11.26394 11.26140 11.25904 11.54495
#> 12 11.49827 11.49461 11.49120 11.48802 11.48506 11.48230 11.47972 11.67088
#> 13 24.78683 24.81285 24.83712 24.85975 24.88085 24.90054 24.91889 22.13410
#> 14 16.69695 16.73906 16.77831 16.81490 16.84904 16.88087 16.91055 17.45032
#> 15 16.93604 16.97783 17.01677 17.05308 17.08695 17.11854 17.14799 17.58580
#> 16 26.27763 26.30162 26.32400 26.34486 26.36432 26.38247 26.39940 26.48014
#> 17 25.66836 25.66677 25.66529 25.66391 25.66263 25.66143 25.66031 26.65245
#> 18 27.43200 27.42801 27.42430 27.42083 27.41760 27.41458 27.41177 29.21878
#> 19 21.22087 21.13542 21.05575 20.98146 20.91217 20.84755 20.78728 19.90731
#> 20 16.36048 16.24456 16.13648 16.03568 15.94168 15.85401 15.77225 15.76167
#> 21 23.86815 23.86901 23.86981 23.87057 23.87127 23.87192 23.87253 23.69801
#>      0.1721   0.1605   0.1497   0.1396   0.1302   0.1214   0.1132   0.1056
#> 1  19.10900 19.10884 19.10872 19.10862 19.10853 19.10846 19.10840 19.10835
#> 2  20.93861 20.93858 20.93856 20.93854 20.93853 20.93851 20.93850 20.93850
#> 3  17.57775 17.57769 17.57763 17.57759 17.57755 17.57752 17.57749 17.57747
#> 4  23.04472 23.04491 23.04507 23.04520 23.04531 23.04540 23.04548 23.04554
#> 5  23.10846 23.10865 23.10880 23.10892 23.10902 23.10911 23.10918 23.10924
#> 6  19.77545 19.77539 19.77533 19.77529 19.77526 19.77523 19.77520 19.77518
#> 7  15.13833 15.13827 15.13823 15.13819 15.13815 15.13812 15.13810 15.13808
#> 8  15.68017 15.68002 15.67990 15.67979 15.67971 15.67964 15.67958 15.67953
#> 9  15.60049 15.60035 15.60024 15.60015 15.60007 15.60000 15.59995 15.59991
#> 10 11.82242 11.82251 11.82259 11.82265 11.82271 11.82275 11.82279 11.82282
#> 11 11.54512 11.54526 11.54538 11.54548 11.54556 11.54562 11.54568 11.54572
#> 12 11.67102 11.67114 11.67124 11.67132 11.67139 11.67145 11.67149 11.67153
#> 13 22.13383 22.13361 22.13342 22.13326 22.13314 22.13303 22.13294 22.13287
#> 14 17.45026 17.45022 17.45018 17.45015 17.45012 17.45010 17.45009 17.45007
#> 15 17.58572 17.58565 17.58560 17.58555 17.58551 17.58548 17.58546 17.58543
#> 16 26.48016 26.48018 26.48019 26.48021 26.48022 26.48023 26.48023 26.48024
#> 17 26.65288 26.65324 26.65354 26.65378 26.65399 26.65416 26.65430 26.65441
#> 18 29.21894 29.21907 29.21918 29.21928 29.21935 29.21941 29.21946 29.21951
#> 19 19.90717 19.90705 19.90696 19.90688 19.90681 19.90676 19.90672 19.90668
#> 20 15.76140 15.76118 15.76099 15.76084 15.76071 15.76061 15.76052 15.76045
#> 21 23.69811 23.69819 23.69826 23.69832 23.69837 23.69840 23.69844 23.69846
#>      0.0985   0.0918   0.0856   0.0799   0.0745   0.0695   0.0648   0.0604
#> 1  19.10831 19.10828 19.10825 19.10823 19.10821 19.10819 19.10818 19.34050
#> 2  20.93849 20.93848 20.93848 20.93847 20.93847 20.93847 20.93847 20.68790
#> 3  17.57745 17.57744 17.57743 17.57742 17.57741 17.57740 17.57740 17.23938
#> 4  23.04559 23.04563 23.04567 23.04570 23.04572 23.04574 23.04576 23.19550
#> 5  23.10929 23.10933 23.10936 23.10939 23.10941 23.10943 23.10945 23.27008
#> 6  19.77517 19.77515 19.77514 19.77513 19.77513 19.77512 19.77511 19.89323
#> 7  15.13807 15.13805 15.13804 15.13803 15.13803 15.13802 15.13802 15.40567
#> 8  15.67949 15.67946 15.67943 15.67941 15.67939 15.67937 15.67936 15.81455
#> 9  15.59987 15.59984 15.59982 15.59980 15.59978 15.59976 15.59975 15.75442
#> 10 11.82284 11.82286 11.82288 11.82290 11.82291 11.82292 11.82293 11.65224
#> 11 11.54576 11.54579 11.54582 11.54584 11.54586 11.54587 11.54588 11.49684
#> 12 11.67156 11.67159 11.67161 11.67163 11.67164 11.67166 11.67167 11.68159
#> 13 22.13281 22.13276 22.13272 22.13268 22.13266 22.13263 22.13261 22.20868
#> 14 17.45006 17.45005 17.45004 17.45004 17.45003 17.45003 17.45002 17.33164
#> 15 17.58542 17.58540 17.58539 17.58538 17.58537 17.58536 17.58536 17.49669
#> 16 26.48025 26.48025 26.48025 26.48026 26.48026 26.48026 26.48026 26.46248
#> 17 26.65451 26.65459 26.65465 26.65471 26.65475 26.65479 26.65482 26.74239
#> 18 29.21954 29.21957 29.21960 29.21962 29.21963 29.21965 29.21966 28.87580
#> 19 19.90665 19.90662 19.90660 19.90658 19.90657 19.90656 19.90655 20.05825
#> 20 15.76039 15.76034 15.76030 15.76026 15.76024 15.76021 15.76019 15.48828
#> 21 23.69849 23.69850 23.69852 23.69853 23.69854 23.69855 23.69856 23.80389
#>      0.0563   0.0525   0.0490   0.0457   0.0426   0.0398   0.0371   0.0346
#> 1  19.34050 19.34050 19.34050 19.34049 19.34049 19.30564 19.30564 19.30563
#> 2  20.68775 20.68762 20.68751 20.68741 20.68733 20.69090 20.69081 20.69073
#> 3  17.23913 17.23892 17.23873 17.23856 17.23842 17.18456 17.18435 17.18415
#> 4  23.19573 23.19594 23.19612 23.19628 23.19642 23.21202 23.21217 23.21231
#> 5  23.27031 23.27051 23.27069 23.27085 23.27099 22.95010 22.94994 22.94979
#> 6  19.89327 19.89331 19.89335 19.89338 19.89341 20.00660 20.00675 20.00689
#> 7  15.40575 15.40582 15.40587 15.40593 15.40597 15.49813 15.49829 15.49844
#> 8  15.81450 15.81444 15.81440 15.81436 15.81432 15.80948 15.80946 15.80945
#> 9  15.75439 15.75435 15.75432 15.75429 15.75427 15.71395 15.71392 15.71388
#> 10 11.65223 11.65222 11.65221 11.65220 11.65219 11.60149 11.60141 11.60134
#> 11 11.49692 11.49700 11.49706 11.49712 11.49717 11.51927 11.51933 11.51938
#> 12 11.68169 11.68178 11.68185 11.68192 11.68198 11.76073 11.76085 11.76096
#> 13 22.20854 22.20842 22.20831 22.20822 22.20814 22.17223 22.17215 22.17208
#> 14 17.33153 17.33142 17.33133 17.33125 17.33118 17.36977 17.36973 17.36969
#> 15 17.49657 17.49646 17.49637 17.49629 17.49622 17.48577 17.48569 17.48561
#> 16 26.46248 26.46249 26.46249 26.46249 26.46249 26.48839 26.48842 26.48844
#> 17 26.74268 26.74294 26.74317 26.74337 26.74354 26.75745 26.75760 26.75774
#> 18 28.87574 28.87568 28.87564 28.87559 28.87556 28.94878 28.94878 28.94877
#> 19 20.05826 20.05826 20.05827 20.05827 20.05828 20.06427 20.06429 20.06431
#> 20 15.48802 15.48779 15.48759 15.48741 15.48725 15.43204 15.43183 15.43163
#> 21 23.80402 23.80413 23.80423 23.80432 23.80439 23.92842 23.92861 23.92879
#>      0.0322   0.0301   0.0280   0.0262   0.0244   0.0227   0.0212   0.0198
#> 1  19.30562 19.30562 19.30561 19.30561 19.30561 19.30560 19.30560 19.30559
#> 2  20.69065 20.69057 20.69050 20.69044 20.69038 20.69032 20.69027 20.69022
#> 3  17.18397 17.18379 17.18363 17.18348 17.18335 17.18322 17.18309 17.18298
#> 4  23.21244 23.21256 23.21267 23.21277 23.21287 23.21296 23.21305 23.21313
#> 5  22.94965 22.94952 22.94940 22.94929 22.94918 22.94908 22.94899 22.94890
#> 6  20.00702 20.00714 20.00725 20.00735 20.00745 20.00754 20.00762 20.00770
#> 7  15.49857 15.49870 15.49881 15.49892 15.49903 15.49912 15.49921 15.49929
#> 8  15.80943 15.80942 15.80941 15.80940 15.80939 15.80938 15.80937 15.80936
#> 9  15.71385 15.71382 15.71379 15.71377 15.71374 15.71372 15.71370 15.71368
#> 10 11.60128 11.60122 11.60116 11.60111 11.60106 11.60101 11.60097 11.60093
#> 11 11.51942 11.51947 11.51951 11.51955 11.51958 11.51961 11.51964 11.51967
#> 12 11.76106 11.76116 11.76125 11.76134 11.76141 11.76149 11.76156 11.76162
#> 13 22.17201 22.17194 22.17188 22.17183 22.17177 22.17172 22.17168 22.17164
#> 14 17.36966 17.36963 17.36960 17.36957 17.36955 17.36952 17.36950 17.36948
#> 15 17.48554 17.48548 17.48541 17.48536 17.48531 17.48526 17.48521 17.48517
#> 16 26.48847 26.48849 26.48851 26.48853 26.48854 26.48856 26.48857 26.48859
#> 17 26.75788 26.75800 26.75812 26.75822 26.75832 26.75842 26.75850 26.75858
#> 18 28.94876 28.94875 28.94875 28.94874 28.94874 28.94873 28.94873 28.94872
#> 19 20.06434 20.06435 20.06437 20.06439 20.06440 20.06442 20.06443 20.06445
#> 20 15.43144 15.43127 15.43110 15.43095 15.43081 15.43068 15.43056 15.43044
#> 21 23.92895 23.92911 23.92925 23.92939 23.92952 23.92963 23.92974 23.92984
#>      0.0185   0.0172   0.0160   0.0150   0.0140   0.0130   0.0121   0.0113
#> 1  19.30559 19.30559 19.30559 19.30558 19.30558 19.30558 19.25992 19.25948
#> 2  20.69017 20.69013 20.69009 20.69005 20.69002 20.68999 20.68421 20.68435
#> 3  17.18288 17.18278 17.18269 17.18260 17.18252 17.18245 17.19285 17.19314
#> 4  23.21320 23.21327 23.21333 23.21339 23.21345 23.21350 23.25491 23.25508
#> 5  22.94882 22.94875 22.94868 22.94862 22.94856 22.94850 22.95194 22.95192
#> 6  20.00778 20.00784 20.00791 20.00797 20.00802 20.00807 20.06184 20.06185
#> 7  15.49937 15.49944 15.49951 15.49957 15.49963 15.49968 15.51489 15.51477
#> 8  15.80935 15.80935 15.80934 15.80933 15.80933 15.80932 15.82107 15.82097
#> 9  15.71367 15.71365 15.71363 15.71362 15.71360 15.71359 15.72012 15.71999
#> 10 11.60089 11.60086 11.60083 11.60080 11.60077 11.60074 11.57813 11.57820
#> 11 11.51970 11.51972 11.51975 11.51977 11.51979 11.51981 11.50149 11.50154
#> 12 11.76168 11.76174 11.76179 11.76184 11.76188 11.76192 11.75171 11.75174
#> 13 22.17160 22.17156 22.17152 22.17149 22.17146 22.17143 22.17659 22.17643
#> 14 17.36946 17.36945 17.36943 17.36941 17.36940 17.36939 17.38815 17.38835
#> 15 17.48513 17.48509 17.48506 17.48502 17.48499 17.48496 17.49984 17.49999
#> 16 26.48860 26.48861 26.48862 26.48863 26.48864 26.48865 26.43718 26.43700
#> 17 26.75866 26.75873 26.75880 26.75886 26.75891 26.75897 26.77191 26.77220
#> 18 28.94872 28.94871 28.94871 28.94871 28.94870 28.94870 28.95899 28.95933
#> 19 20.06446 20.06447 20.06448 20.06449 20.06450 20.06450 20.07556 20.07536
#> 20 15.43034 15.43024 15.43014 15.43006 15.42998 15.42990 15.42402 15.42389
#> 21 23.92994 23.93003 23.93011 23.93019 23.93026 23.93033 23.87469 23.87442
#>      0.0106   0.0098   0.0092   0.0086   0.0080   0.0074   0.0069   0.0065
#> 1  19.25913 19.25885 19.25864 19.25846 19.25833 19.25823 19.25815 19.25809
#> 2  20.68447 20.68458 20.68467 20.68475 20.68482 20.68487 20.68492 20.68497
#> 3  17.19339 17.19360 17.19378 17.19393 17.19406 17.19416 17.19426 17.19434
#> 4  23.25520 23.25528 23.25533 23.25536 23.25538 23.25538 23.25538 23.25537
#> 5  22.95191 22.95190 22.95190 22.95190 22.95190 22.95191 22.95191 22.95192
#> 6  20.06183 20.06180 20.06176 20.06172 20.06167 20.06162 20.06158 20.06153
#> 7  15.51467 15.51457 15.51448 15.51440 15.51433 15.51426 15.51421 15.51415
#> 8  15.82088 15.82081 15.82074 15.82069 15.82064 15.82061 15.82058 15.82055
#> 9  15.71988 15.71978 15.71971 15.71965 15.71960 15.71956 15.71952 15.71950
#> 10 11.57828 11.57834 11.57841 11.57847 11.57852 11.57856 11.57861 11.57864
#> 11 11.50158 11.50162 11.50165 11.50167 11.50170 11.50172 11.50173 11.50174
#> 12 11.75176 11.75178 11.75179 11.75180 11.75180 11.75180 11.75180 11.75180
#> 13 22.17630 22.17619 22.17610 22.17603 22.17598 22.17594 22.17591 22.17588
#> 14 17.38851 17.38864 17.38874 17.38882 17.38889 17.38894 17.38898 17.38902
#> 15 17.50011 17.50021 17.50029 17.50035 17.50041 17.50045 17.50049 17.50052
#> 16 26.43688 26.43679 26.43672 26.43668 26.43665 26.43662 26.43661 26.43660
#> 17 26.77243 26.77262 26.77276 26.77286 26.77294 26.77300 26.77304 26.77307
#> 18 28.95960 28.95983 28.96002 28.96017 28.96029 28.96039 28.96047 28.96053
#> 19 20.07519 20.07504 20.07492 20.07482 20.07473 20.07466 20.07461 20.07456
#> 20 15.42379 15.42372 15.42367 15.42364 15.42363 15.42362 15.42362 15.42363
#> 21 23.87422 23.87406 23.87393 23.87383 23.87375 23.87368 23.87363 23.87359
#>      0.0060   0.0056   0.0053   0.0049   0.0046
#> 1  19.25804 19.25801 19.25799 19.25797 19.25796
#> 2  20.68501 20.68504 20.68507 20.68509 20.68511
#> 3  17.19441 17.19446 17.19452 17.19456 17.19460
#> 4  23.25535 23.25533 23.25531 23.25529 23.25527
#> 5  22.95194 22.95195 22.95196 22.95198 22.95199
#> 6  20.06149 20.06145 20.06141 20.06137 20.06134
#> 7  15.51411 15.51407 15.51403 15.51400 15.51397
#> 8  15.82053 15.82051 15.82050 15.82049 15.82048
#> 9  15.71948 15.71946 15.71945 15.71944 15.71944
#> 10 11.57867 11.57870 11.57873 11.57875 11.57877
#> 11 11.50175 11.50176 11.50176 11.50177 11.50177
#> 12 11.75179 11.75178 11.75177 11.75177 11.75176
#> 13 22.17586 22.17585 22.17584 22.17584 22.17584
#> 14 17.38905 17.38907 17.38908 17.38910 17.38911
#> 15 17.50054 17.50056 17.50058 17.50060 17.50061
#> 16 26.43660 26.43660 26.43660 26.43660 26.43661
#> 17 26.77309 26.77310 26.77310 26.77310 26.77309
#> 18 28.96058 28.96062 28.96065 28.96067 28.96069
#> 19 20.07452 20.07449 20.07446 20.07444 20.07442
#> 20 15.42364 15.42366 15.42368 15.42370 15.42372
#> 21 23.87355 23.87352 23.87349 23.87347 23.87345
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.4 18.7 24.4 22.8 19.2 16.4 17.3 15.2 10.4 10.4 14.7 21.5 15.5 15.2
#> [16] 27.3 26.0 30.4 19.7 15.0 21.4
#> 
#> $X
#>               am        carb        cyl       disp       drat       gear
#>  [1,]  1.4142136  0.62526247 -0.1608169 -0.6636680  0.9042522  0.4291975
#>  [2,] -0.7071068 -1.16526187 -0.1608169  0.1512656 -0.8834886 -0.8583951
#>  [3,] -0.7071068 -0.56842042  0.9649013  0.9994618 -0.7308766 -0.8583951
#>  [4,] -0.7071068 -0.56842042 -1.2865350 -0.7742661  0.4464161  0.4291975
#>  [5,] -0.7071068 -0.56842042 -1.2865350 -0.8233284  0.9478556  0.4291975
#>  [6,] -0.7071068  0.62526247 -0.1608169 -0.6004691  0.9478556  0.4291975
#>  [7,] -0.7071068  0.02842102  0.9649013  0.2992842 -0.9052904 -0.8583951
#>  [8,] -0.7071068  0.02842102  0.9649013  0.2992842 -0.9052904 -0.8583951
#>  [9,] -0.7071068  0.02842102  0.9649013  0.2992842 -0.9052904 -0.8583951
#> [10,] -0.7071068  0.62526247  0.9649013  1.9308145 -1.2105144 -0.8583951
#> [11,] -0.7071068  0.62526247  0.9649013  1.8310267 -1.0579024 -0.8583951
#> [12,] -0.7071068  0.62526247  0.9649013  1.6647137 -0.5564629 -0.8583951
#> [13,] -0.7071068 -1.16526187 -1.2865350 -0.9954624  0.4682178 -0.8583951
#> [14,] -0.7071068 -0.56842042  0.9649013  0.6502046 -1.5811436 -0.8583951
#> [15,] -0.7071068 -0.56842042  0.9649013  0.5337855 -0.7308766 -0.8583951
#> [16,]  1.4142136 -1.16526187 -1.2865350 -1.3372355  1.2966831  0.4291975
#> [17,]  1.4142136 -0.56842042 -1.2865350 -0.9937992  2.0597432  1.7167902
#> [18,]  1.4142136 -0.56842042 -1.2865350 -1.2033536  0.6208299  1.7167902
#> [19,]  1.4142136  1.81894535 -0.1608169 -0.7884027  0.2938041  1.7167902
#> [20,]  1.4142136  3.01262824  0.9649013  0.5088385  0.1193904  1.7167902
#> [21,]  1.4142136 -0.56842042 -1.2865350 -0.9879783  1.3620883  0.4291975
#>                  hp        qsec         vs          wt
#>  [1,] -0.6373227849 -0.52899028 -0.7844645 -0.50280220
#>  [2,] -0.6373227849  0.89642256  1.2747549 -0.16609167
#>  [3,]  0.3970952905 -0.52899028 -0.7844645  0.05673148
#>  [4,] -1.4012007482  1.22627016  1.2747549 -0.19084979
#>  [5,] -0.8760346484  2.93440951  1.2747549 -0.23046280
#>  [6,] -0.4304391698  0.22494709  1.2747549  0.05673148
#>  [7,]  0.4766659116 -0.30516513 -0.7844645  0.68063629
#>  [8,]  0.4766659116 -0.18736241 -0.7844645  0.34392576
#>  [9,]  0.4766659116  0.04824302 -0.7844645  0.39344201
#> [10,]  0.8745190176  0.03646274 -0.7844645  1.84921989
#> [11,]  1.0336602599 -0.05777943 -0.7844645  2.02153646
#> [12,]  1.2723721235 -0.29338485 -0.7844645  1.94330078
#> [13,] -0.8442064000  1.23216029  1.2747549 -0.90883548
#> [14,] -0.0007578154 -0.61734232 -0.7844645  0.13595749
#> [15,] -0.0007578154 -0.36406648 -0.7844645  0.05177985
#> [16,] -1.3375442513  0.57835523  1.2747549 -1.43370778
#> [17,] -0.9396911454 -0.71747463 -0.7844645 -1.23069114
#> [18,] -0.5895804122 -0.59967191  1.2747549 -1.85162497
#> [19,]  0.3970952905 -1.42429091 -0.7844645 -0.60678633
#> [20,]  2.9433551683 -1.95440313 -0.7844645  0.18547374
#> [21,] -0.6532369091  0.40165116  1.2747549 -0.59688308
#> attr(,"center")
#>  [1]   0.3333333   2.9523810   6.2857143 239.8095238   3.4852381   3.6666667
#>  [7] 150.0476190  17.9180952   0.3809524   3.3827143
#> attr(,"scale")
#>  [1]   0.4714045   1.6754869   1.7766436 120.2551965   0.4586795   0.7766432
#>  [7]  62.8372624   1.6977538   0.4856209   1.0097694
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 39
#> 
#> $lambda.min
#> [1] 0.3224292
#> 
#> $null.dev
#> [1] 25.9946
#> 
#> $Bias
#> [1] 2.101268
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 13.54732