Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 32.623277 31.154968 29.274511 27.379922 25.385599 23.154271 21.257222
#>   [8] 19.647970 18.286286 17.137365 16.171109 15.362658 14.726395 14.211397
#>  [15] 13.764727 13.355933 12.868868 12.410162 11.916231 11.351113 11.018833
#>  [22] 10.816752 10.662619 10.548733 10.468491 10.416243 10.387161 10.377124
#>  [29] 10.335800 10.291177 10.315133 10.492821  7.318305  7.719681  7.785907
#>  [36]  7.872709  7.954197  9.079280  9.080040  9.019738  9.054424  9.219604
#>  [43]  9.251953 10.094774 10.121949 10.144621 13.385464 13.544060 13.284008
#>  [50] 13.284844 13.285708 13.438978 13.633690 18.448053 18.204806 18.205718
#>  [57] 18.130227 17.955901 19.256908 19.258313 19.210328 19.056567 18.868051
#>  [64] 19.293990 19.529668 19.530996 19.496962 19.498250 19.499491 19.408130
#>  [71] 19.409224 19.410271 19.644644 19.645427 19.533750 19.534324 19.269381
#>  [78] 19.268858 19.268364 19.267899 19.267462 19.267050 19.266665 19.266303
#>  [85] 19.265964 19.265648 19.265353 19.265077 19.264821 19.264582 19.264361
#>  [92] 19.264156 19.263966 19.263848 19.279014 19.278937 19.278873 19.278819
#>  [99] 19.278775 19.278741
#> 
#> $cvse
#>   [1] 11.330658 11.380053 11.143416 10.851648 10.359029  9.561586  8.861329
#>   [8]  8.244253  7.698510  7.214112  6.782659  6.396995  6.048353  5.735331
#>  [15]  5.455625  5.207836  4.997161  4.807725  4.630186  4.459997  4.414528
#>  [22]  4.425828  4.439867  4.455525  4.472012  4.488783  4.505461  4.521798
#>  [29]  4.526765  4.529089  4.532158  4.540293  2.688938  2.705747  2.703750
#>  [36]  2.701314  2.709504  2.698435  2.708735  2.735592  2.762396  2.823422
#>  [43]  2.846651  2.902779  2.921076  2.868683  3.644148  3.700289  3.711429
#>  [50]  3.711547  3.711682  3.765158  3.774368  6.489321  6.470099  6.470587
#>  [57]  6.469717  6.481455  7.666183  7.667151  7.672745  7.652094  7.673418
#>  [64]  7.716833  7.750172  7.751030  7.755479  7.756261  7.757009  7.760178
#>  [71]  7.760867  7.761528  7.809644  7.810212  7.815045  7.815570  7.757750
#>  [78]  7.758034  7.758306  7.758567  7.758818  7.759059  7.759290  7.759513
#>  [85]  7.759727  7.759934  7.760132  7.760323  7.760507  7.760685  7.760856
#>  [92]  7.761020  7.761179  7.761330  7.760740  7.760878  7.761010  7.761138
#>  [99]  7.761261  7.761379
#> 
#> $fold
#>  [1]  6  3  1  8  9 10  7  4  2  3  2  4  7  5 10  6  1  5  8  9  1
#> 
#> $lambda
#>   [1] 4.820575904 4.495685222 4.192691085 3.910117738 3.646588889 3.400821003
#>   [7] 3.171617049 2.957860675 2.758510765 2.572596372 2.399211987 2.237513129
#>  [13] 2.086712233 1.946074812 1.814915883 1.692596627 1.578521279 1.472134228
#>  [19] 1.372917308 1.280387277 1.194093459 1.113615557 1.038561595 0.968566020
#>  [25] 0.903287912 0.842409330 0.785633760 0.732684674 0.683304179 0.637251765
#>  [31] 0.594303129 0.554249087 0.516894553 0.482057590 0.449568522 0.419269108
#>  [37] 0.391011774 0.364658889 0.340082100 0.317161705 0.295786068 0.275851077
#>  [43] 0.257259637 0.239921199 0.223751313 0.208671223 0.194607481 0.181491588
#>  [49] 0.169259663 0.157852128 0.147213423 0.137291731 0.128038728 0.119409346
#>  [55] 0.111361556 0.103856160 0.096856602 0.090328791 0.084240933 0.078563376
#>  [61] 0.073268467 0.068330418 0.063725176 0.059430313 0.055424909 0.051689455
#>  [67] 0.048205759 0.044956852 0.041926911 0.039101177 0.036465889 0.034008210
#>  [73] 0.031716170 0.029578607 0.027585108 0.025725964 0.023992120 0.022375131
#>  [79] 0.020867122 0.019460748 0.018149159 0.016925966 0.015785213 0.014721342
#>  [85] 0.013729173 0.012803873 0.011940935 0.011136156 0.010385616 0.009685660
#>  [91] 0.009032879 0.008424093 0.007856338 0.007326847 0.006833042 0.006372518
#>  [97] 0.005943031 0.005542491 0.005168946 0.004820576
#> 
#> $fit
#> $beta
#>               4.8206     4.4957    4.1927    3.9101    3.6466    3.4008
#> (Intercept) 19.56667 21.4914878 23.286583 24.960694 26.521976 27.978032
#> am           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000 -0.6014171 -1.162301 -1.685383 -2.173211 -2.628161
#>                3.1716    2.9579    2.7585    2.5726    2.3992    2.2375
#> (Intercept) 29.335956 30.602359 31.783412 32.884865 33.912084 34.870072
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -3.052449 -3.448141 -3.817165 -4.161318 -4.482276 -4.781603
#>                2.0867    1.9461    1.8149      1.6926     1.5785     1.4721
#> (Intercept) 35.763495 36.596704 37.373758 37.48571010 36.5284240 34.3650391
#> am           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> carb         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> cyl          0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> disp         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> qsec         0.000000  0.000000  0.000000  0.02986232  0.1063629  0.2170977
#> vs           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> wt          -5.060756 -5.321095 -5.563888 -5.76283298 -5.8837674 -5.8158217
#>                1.3729     1.2804     1.1941     1.1136    1.0386     0.9686
#> (Intercept) 32.346606 30.4642086 28.7086784 27.0714650 25.544594 24.1206296
#> am           0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> carb         0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> cyl          0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> disp         0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> drat         0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> gear         0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> hp           0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> qsec         0.320413  0.4167651  0.5066235  0.5904257  0.668580  0.7414669
#> vs           0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> wt          -5.752429 -5.6933079 -5.6381717 -5.5867516 -5.538797 -5.4940743
#>                 0.9033     0.8424     0.7856     0.7327    0.6833    0.6373
#> (Intercept) 22.7926354 21.5541435 20.3991218 19.3219447 18.317366 17.380492
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> qsec         0.8094414  0.8728347  0.9319555  0.9870918  1.038512  1.086467
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> wt          -5.4523658 -5.4134684 -5.3771925 -5.3433614 -5.311811 -5.282386
#>                0.5943    0.5542    0.5169    0.4821    0.4496    0.4193
#> (Intercept) 16.931490 16.931198 16.931172 16.931169 16.931169 16.931169
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         1.109449  1.109464  1.109466  1.109466  1.109466  1.109466
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -5.268284 -5.268275 -5.268274 -5.268274 -5.268274 -5.268274
#>                 0.3910    0.3647    0.3401    0.3172    0.2958     0.2759
#> (Intercept) 15.9966463 11.750083  8.049075 -1.480063 -1.484822 -13.064030
#> am           0.2224633  1.231647  2.111129  2.554781  2.555528   1.957467
#> carb         0.0000000  0.000000  0.000000  0.000000  0.000000   0.000000
#> cyl          0.0000000  0.000000  0.000000  0.000000  0.000000   0.000000
#> disp         0.0000000  0.000000  0.000000  0.000000  0.000000   0.000000
#> drat         0.0000000  0.000000  0.000000  0.000000  0.000000   1.826229
#> gear         0.0000000  0.000000  0.000000  0.000000  0.000000   0.000000
#> hp           0.0000000  0.000000  0.000000  0.000000  0.000000   0.000000
#> qsec         1.1386799  1.271575  1.387402  2.050604  2.050802   2.284437
#> vs           0.0000000  0.000000  0.000000 -2.817454 -2.817679  -3.943910
#> wt          -5.1631652 -4.686123 -4.270384 -4.609943 -4.609603  -4.123780
#>                 0.2573     0.2399     0.2238     0.2087     0.1946     0.1815
#> (Intercept) -13.071200 -13.076965 -13.081593 -13.085305 -13.088281 -13.090663
#> am            1.957677   1.957845   1.957979   1.958086   1.958172   1.958240
#> carb          0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> cyl           0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> disp          0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> drat          1.826757   1.827183   1.827526   1.827802   1.828023   1.828201
#> gear          0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> hp            0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> qsec          2.284698   2.284908   2.285075   2.285209   2.285317   2.285402
#> vs           -3.944630  -3.945206  -3.945666  -3.946034  -3.946328  -3.946563
#> wt           -4.123500  -4.123273  -4.123091  -4.122944  -4.122826  -4.122731
#>                 0.1693     0.1579     0.1472      0.1373      0.1280
#> (Intercept) -13.092571 -13.094097 -13.095317 -33.7718613 -33.7752383
#> am            1.958294   1.958337   1.958371   1.5798825   1.5801631
#> carb          0.000000   0.000000   0.000000   0.0000000   0.0000000
#> cyl           0.000000   0.000000   0.000000   0.8942216   0.8944877
#> disp          0.000000   0.000000   0.000000   0.0000000   0.0000000
#> drat          1.828344   1.828459   1.828550   1.9729699   1.9727634
#> gear          0.000000   0.000000   0.000000   1.1805573   1.1806895
#> hp            0.000000   0.000000   0.000000   0.0000000   0.0000000
#> qsec          2.285471   2.285526   2.285570   3.0195161   3.0196679
#> vs           -3.946752  -3.946902  -3.947022  -4.2705431  -4.2704200
#> wt           -4.122656  -4.122595  -4.122546  -4.8937261  -4.8939963
#>                  0.1194      0.1114      0.1039      0.0969      0.0903
#> (Intercept) -33.7780821 -33.7804734 -33.7824812 -33.7841643 -33.7855729
#> am            1.5804051   1.5806136   1.5807930   1.5809471   1.5810794
#> carb          0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> cyl           0.8947138   0.8949057   0.8950684   0.8952061   0.8953226
#> disp          0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> drat          1.9725859   1.9724333   1.9723023   1.9721899   1.9720937
#> gear          1.1808002   1.1808928   1.1809701   1.1810344   1.1810880
#> hp            0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> qsec          3.0197961   3.0199042   3.0199952   3.0200717   3.0201359
#> vs           -4.2703123  -4.2702183  -4.2701365  -4.2700652  -4.2700033
#> wt           -4.8942260  -4.8944209  -4.8945861  -4.8947260  -4.8948443
#>                   0.0842       0.0786       0.0733       0.0683       0.0637
#> (Intercept) -28.10018489 -28.09732753 -28.09474862 -28.09242609 -28.09033921
#> am            1.78216569   1.78208030   1.78200112   1.78192782   1.78186008
#> carb          0.70938515   0.70958269   0.70976449   0.70993171   0.71008546
#> cyl           0.66420424   0.66391683   0.66365288   0.66341077   0.66318898
#> disp          0.01673901   0.01674266   0.01674600   0.01674904   0.01675181
#> drat          1.60881412   1.60868867   1.60857799   1.60848057   1.60839499
#> gear          0.86608746   0.86585358   0.86563786   0.86543918   0.86525641
#> hp           -0.01958185  -0.01958307  -0.01958418  -0.01958519  -0.01958610
#> qsec          2.97613289   2.97614255   2.97615313   2.97616452   2.97617661
#> vs           -3.84017599  -3.84040401  -3.84061980  -3.84082363  -3.84101581
#> wt           -6.21578213  -6.21609939  -6.21638859  -6.21665225  -6.21689268
#>                   0.0594       0.0554       0.0517       0.0482       0.0450
#> (Intercept) -28.08846854 -28.08679588 -28.08530425 -28.08397777 -28.08280168
#> am            1.78179756   1.78173994   1.78168690   1.78163813   1.78159333
#> carb          0.71022675   0.71035652   0.71047567   0.71058502   0.71068533
#> cyl           0.66298604   0.66280057   0.66263126   0.66247688   0.66233627
#> disp          0.01675434   0.01675665   0.01675876   0.01676068   0.01676243
#> drat          1.60831999   1.60825438   1.60819713   1.60814728   1.60810396
#> gear          0.86508853   0.86493452   0.86479343   0.86466435   0.86454643
#> hp           -0.01958693  -0.01958769  -0.01958837  -0.01958899  -0.01958955
#> qsec          2.97618930   2.97620250   2.97621612   2.97623007   2.97624427
#> vs           -3.84119671  -3.84136669  -3.84152618  -3.84167561  -3.84181541
#> wt           -6.21711201  -6.21731217  -6.21749494  -6.21766193  -6.21781462
#>                   0.0419       0.0391       0.0365       0.0340       0.0317
#> (Intercept) -28.08176229 -28.08084686 -28.08004363 -28.07934171 -28.07873107
#> am            1.78155221   1.78151449   1.78147992   1.78144823   1.78141919
#> carb          0.71077732   0.71086164   0.71093890   0.71100968   0.71107449
#> cyl           0.66220833   0.66209205   0.66198647   0.66189070   0.66180391
#> disp          0.01676403   0.01676548   0.01676681   0.01676803   0.01676914
#> drat          1.60806640   1.60803390   1.60800584   1.60798165   1.60796084
#> gear          0.86443886   0.86434086   0.86425172   0.86417076   0.86409733
#> hp           -0.01959006  -0.01959052  -0.01959094  -0.01959132  -0.01959166
#> qsec          2.97625865   2.97627315   2.97628770   2.97630223   2.97631671
#> vs           -3.84194604  -3.84206795  -3.84218159  -3.84228740  -3.84238583
#> wt           -6.21795433  -6.21808228  -6.21819959  -6.21830723  -6.21840613
#>                   0.0296       0.0276       0.0257       0.0240       0.0224
#> (Intercept) -28.07820249 -28.07774747 -28.07735824 -28.07702769 -28.07674931
#> am            1.78139259   1.78136822   1.78134587   1.78132538   1.78130657
#> carb          0.71113382   0.71118812   0.71123781   0.71128326   0.71132482
#> cyl           0.66172532   0.66165422   0.66158995   0.66153189   0.66147947
#> disp          0.01677015   0.01677108   0.01677193   0.01677271   0.01677342
#> drat          1.60794296   1.60792762   1.60791446   1.60790319   1.60789352
#> gear          0.86403083   0.86397072   0.86391645   0.86386754   0.86382354
#> hp           -0.01959196  -0.01959224  -0.01959249  -0.01959271  -0.01959290
#> qsec          2.97633108   2.97634529   2.97635932   2.97637312   2.97638665
#> vs           -3.84247730  -3.84256223  -3.84264101  -3.84271405  -3.84278171
#> wt           -6.21849709  -6.21858085  -6.21865808  -6.21872939  -6.21879530
#>                   0.0209       0.0195       0.0181       0.0169       0.0158
#> (Intercept) -28.07651718 -28.07632592 -28.07617064 -28.07604691 -28.07595074
#> am            1.78128929   1.78127340   1.78125877   1.78124528   1.78123282
#> carb          0.71136283   0.71139757   0.71142933   0.71145836   0.71148488
#> cyl           0.66143218   0.66138953   0.66135109   0.66131645   0.66128524
#> disp          0.01677407   0.01677467   0.01677522   0.01677573   0.01677619
#> drat          1.60788523   1.60787810   1.60787196   1.60786665   1.60786203
#> gear          0.86378402   0.86374858   0.86371687   0.86368854   0.86366328
#> hp           -0.01959308  -0.01959324  -0.01959337  -0.01959349  -0.01959360
#> qsec          2.97639991   2.97641285   2.97642547   2.97643774   2.97644966
#> vs           -3.84284436  -3.84290232  -3.84295592  -3.84300548  -3.84305128
#> wt           -6.21885630  -6.21891285  -6.21896531  -6.21901406  -6.21905941
#>                   0.0147       0.0137       0.0128       0.0119       0.0111
#> (Intercept) -28.07587853 -28.07582704 -28.07579338 -28.07577494 -28.07576942
#> am            1.78122129   1.78121060   1.78120066   1.78119141   1.78118278
#> carb          0.71150912   0.71153126   0.71155148   0.71156996   0.71158683
#> cyl           0.66125713   0.66123180   0.66120898   0.66118841   0.66116987
#> disp          0.01677662   0.01677701   0.01677738   0.01677771   0.01677802
#> drat          1.60785800   1.60785444   1.60785128   1.60784845   1.60784588
#> gear          0.86364081   0.86362085   0.86360317   0.86358754   0.86357375
#> hp           -0.01959369  -0.01959377  -0.01959384  -0.01959389  -0.01959394
#> qsec          2.97646120   2.97647236   2.97648314   2.97649353   2.97650353
#> vs           -3.84309359  -3.84313267  -3.84316877  -3.84320211  -3.84323290
#> wt           -6.21910165  -6.21914102  -6.21917777  -6.21921210  -6.21924419
#>                   0.0104       0.0097       0.0090       0.0084       0.0079
#> (Intercept) -28.07577476 -28.07578915 -28.07581097 -28.07583881 -28.07587145
#> am            1.78117471   1.78116715   1.78116004   1.78115334   1.78114702
#> carb          0.71160225   0.71161632   0.71162918   0.71164092   0.71165164
#> cyl           0.66115315   0.66113805   0.66112440   0.66111206   0.66110089
#> disp          0.01677831   0.01677857   0.01677882   0.01677905   0.01677926
#> drat          1.60784353   1.60784136   1.60783932   1.60783741   1.60783559
#> gear          0.86356162   0.86355098   0.86354167   0.86353354   0.86352648
#> hp           -0.01959398  -0.01959401  -0.01959403  -0.01959405  -0.01959406
#> qsec          2.97651314   2.97652236   2.97653119   2.97653964   2.97654772
#> vs           -3.84326134  -3.84328761  -3.84331189  -3.84333433  -3.84335508
#> wt           -6.21927422  -6.21930234  -6.21932868  -6.21935337  -6.21937653
#>                   0.0073       0.0068       0.0064       0.0059       0.0055
#> (Intercept) -28.07590780 -28.07594692 -28.07598801 -28.07603038 -28.07607343
#> am            1.78114105   1.78113538   1.78113001   1.78112489   1.78112002
#> carb          0.71166143   0.71167037   0.71167854   0.71168599   0.71169279
#> cyl           0.66109076   0.66108156   0.66107319   0.66106557   0.66105861
#> disp          0.01677945   0.01677963   0.01677980   0.01677995   0.01678010
#> drat          1.60783384   1.60783216   1.60783053   1.60782896   1.60782742
#> gear          0.86352035   0.86351506   0.86351051   0.86350661   0.86350329
#> hp           -0.01959406  -0.01959406  -0.01959406  -0.01959405  -0.01959404
#> qsec          2.97655542   2.97656277   2.97656976   2.97657641   2.97658272
#> vs           -3.84337427  -3.84339204  -3.84340849  -3.84342373  -3.84343787
#> wt           -6.21939825  -6.21941862  -6.21943774  -6.21945568  -6.21947251
#>                   0.0052       0.0048
#> (Intercept) -28.07611667 -28.07615969
#> am            1.78111538   1.78111094
#> carb          0.71169901   0.71170468
#> cyl           0.66105224   0.66104641
#> disp          0.01678023   0.01678036
#> drat          1.60782593   1.60782447
#> gear          0.86350047   0.86349809
#> hp           -0.01959403  -0.01959401
#> qsec          2.97658871   2.97659439
#> vs           -3.84345099  -3.84346317
#> wt           -6.21948830  -6.21950312
#> 
#> $iter
#>   [1]   0   3   3   2   2   2   2   2   2   2   2   2   2   2   2   6   5   4
#>  [19]   4   4   4   4   4   4   4   4   4   4   4   4   3   1   1   1   1   1
#>  [37]  40  59  38  43   1  41   1   1   1   1   1   1   1   1   1  72   1   1
#>  [55]   1   1   1   1 232   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.820575904 4.495685222 4.192691085 3.910117738 3.646588889 3.400821003
#>   [7] 3.171617049 2.957860675 2.758510765 2.572596372 2.399211987 2.237513129
#>  [13] 2.086712233 1.946074812 1.814915883 1.692596627 1.578521279 1.472134228
#>  [19] 1.372917308 1.280387277 1.194093459 1.113615557 1.038561595 0.968566020
#>  [25] 0.903287912 0.842409330 0.785633760 0.732684674 0.683304179 0.637251765
#>  [31] 0.594303129 0.554249087 0.516894553 0.482057590 0.449568522 0.419269108
#>  [37] 0.391011774 0.364658889 0.340082100 0.317161705 0.295786068 0.275851077
#>  [43] 0.257259637 0.239921199 0.223751313 0.208671223 0.194607481 0.181491588
#>  [49] 0.169259663 0.157852128 0.147213423 0.137291731 0.128038728 0.119409346
#>  [55] 0.111361556 0.103856160 0.096856602 0.090328791 0.084240933 0.078563376
#>  [61] 0.073268467 0.068330418 0.063725176 0.059430313 0.055424909 0.051689455
#>  [67] 0.048205759 0.044956852 0.041926911 0.039101177 0.036465889 0.034008210
#>  [73] 0.031716170 0.029578607 0.027585108 0.025725964 0.023992120 0.022375131
#>  [79] 0.020867122 0.019460748 0.018149159 0.016925966 0.015785213 0.014721342
#>  [85] 0.013729173 0.012803873 0.011940935 0.011136156 0.010385616 0.009685660
#>  [91] 0.009032879 0.008424093 0.007856338 0.007326847 0.006833042 0.006372518
#>  [97] 0.005943031 0.005542491 0.005168946 0.004820576
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 36
#> 
#> $loss
#>   [1] 648.72667 555.04600 476.66819 411.39104 357.31328 312.79439 276.41983
#>   [8] 246.97068 223.39747 204.79736 190.39435 179.52218 171.60943 166.16662
#>  [15] 162.77502 157.65905 148.83931 137.73234 128.53981 120.98486 114.82452
#>  [22] 109.84948 105.87953 102.75971 100.35683  98.55658  97.26094  96.38599
#>  [29]  95.85994  95.62153  95.59354  95.59354  95.59354  95.59354  95.59354
#>  [36]  95.59354  93.88617  88.52354  87.03425  77.62540  77.62539  71.54363
#>  [43]  71.54362  71.54361  71.54361  71.54361  71.54361  71.54360  71.54360
#>  [50]  71.54360  71.54360  65.02527  65.02526  65.02526  65.02525  65.02525
#>  [57]  65.02525  65.02525  59.32751  59.32750  59.32749  59.32748  59.32747
#>  [64]  59.32746  59.32745  59.32745  59.32745  59.32744  59.32744  59.32744
#>  [71]  59.32744  59.32743  59.32743  59.32743  59.32743  59.32743  59.32743
#>  [78]  59.32743  59.32743  59.32743  59.32743  59.32743  59.32743  59.32743
#>  [85]  59.32743  59.32743  59.32743  59.32743  59.32743  59.32743  59.32743
#>  [92]  59.32743  59.32743  59.32743  59.32743  59.32743  59.32743  59.32743
#>  [99]  59.32743  59.32743
#> 
#> $linear.predictors
#>      4.8206   4.4957   4.1927   3.9101   3.6466   3.4008   3.1716   2.9579
#> 1  19.56667 20.09620 20.59004 21.05061 21.48013 21.88070 22.25427 22.60267
#> 2  19.56667 19.55793 19.54979 19.54219 19.53510 19.52850 19.52233 19.51659
#> 3  19.56667 19.34443 19.13717 18.94388 18.76361 18.59550 18.43871 18.29250
#> 4  19.56667 19.57297 19.57884 19.58432 19.58943 19.59420 19.59864 19.60279
#> 5  19.56667 19.42261 19.28827 19.16298 19.04613 18.93716 18.83553 18.74075
#> 6  19.56667 19.42261 19.28827 19.16298 19.04613 18.93716 18.83553 18.74075
#> 7  19.56667 19.04372 18.55602 18.10119 17.67701 17.28142 16.91249 16.56843
#> 8  19.56667 19.24820 18.95120 18.67422 18.41590 18.17499 17.95032 17.74079
#> 9  19.56667 18.33405 17.18450 16.11243 15.11262 14.18019 13.31060 12.49962
#> 10 19.56667 20.52020 21.40947 22.23880 23.01224 23.73355 24.40625 25.03361
#> 11 19.56667 20.38789 21.15376 21.86802 22.53413 23.15536 23.73471 24.27502
#> 12 19.56667 20.00899 20.42151 20.80623 21.16501 21.49962 21.81167 22.10269
#> 13 19.56667 19.37450 19.19528 19.02815 18.87227 18.72691 18.59134 18.46490
#> 14 19.56667 19.42562 19.29408 19.17140 19.05700 18.95030 18.85079 18.75800
#> 15 19.56667 19.18205 18.82335 18.48882 18.17685 17.88590 17.61455 17.36150
#> 16 19.56667 19.17904 18.81754 18.48040 18.16598 17.87275 17.59929 17.34426
#> 17 19.56667 20.20446 20.79926 21.35397 21.87130 22.35377 22.80372 23.22334
#> 18 19.56667 19.58500 19.60209 19.61803 19.63290 19.64676 19.65969 19.67175
#> 19 19.56667 19.82556 20.06701 20.29218 20.50218 20.69803 20.88067 21.05101
#> 20 19.56667 19.34443 19.13717 18.94388 18.76361 18.59550 18.43871 18.29250
#> 21 19.56667 19.81955 20.05539 20.27533 20.48045 20.67175 20.85015 21.01653
#>      2.7585   2.5726   2.3992    2.2375    2.0867    1.9461    1.8149    1.6926
#> 1  22.92759 23.23061 23.51320 23.776754 24.022542 24.251764 24.465538 24.671675
#> 2  19.51123 19.50623 19.50157 19.497220 19.493165 19.489384 19.485858 19.538726
#> 3  18.15613 18.02896 17.91036 17.799751 17.696597 17.600395 17.510678 17.385416
#> 4  19.60666 19.61026 19.61362 19.616760 19.619684 19.622411 19.624955 19.699519
#> 5  18.65236 18.56993 18.49305 18.421359 18.354495 18.292138 18.233983 18.208045
#> 6  18.65236 18.56993 18.49305 18.421359 18.354495 18.292138 18.233983 18.225963
#> 7  16.24755 15.94830 15.66922 15.408949 15.166219 14.939848 14.728733 14.550584
#> 8  17.54539 17.36315 17.19319 17.034694 16.886876 16.749020 16.620455 16.515920
#> 9  11.74330 11.03795 10.38014  9.766658  9.194527  8.660956  8.163346  7.767761
#> 10 25.61869 26.16434 26.67321 27.147784 27.590375 28.003136 28.388079 28.731785
#> 11 24.77891 25.24885 25.68711 26.095831 26.477008 26.832495 27.164023 27.505172
#> 12 22.37410 22.62722 22.86327 23.083422 23.288732 23.480205 23.658774 23.877872
#> 13 18.34699 18.23703 18.13447 18.038831 17.949635 17.866450 17.788872 17.704315
#> 14 18.67145 18.59074 18.51547 18.445267 18.379799 18.318743 18.261802 18.206997
#> 15 17.12550 16.90540 16.70014 16.508718 16.330193 16.163700 16.008428 15.816610
#> 16 17.10641 16.88460 16.67773 16.484810 16.304889 16.137094 15.980608 15.836770
#> 17 23.61468 23.97965 24.32001 24.637443 24.933478 25.209561 25.467038 25.651948
#> 18 19.68300 19.69349 19.70327 19.712392 19.720899 19.728833 19.736233 19.650533
#> 19 21.20987 21.35801 21.49618 21.625033 21.745202 21.857271 21.961788 21.985529
#> 20 18.15613 18.02896 17.91036 17.799751 17.696597 17.600395 17.510678 17.348386
#> 21 21.17169 21.31640 21.45136 21.577217 21.694594 21.804060 21.906149 22.020474
#>      1.5785    1.4721    1.3729   1.2804    1.1941    1.1136    1.0386
#> 1  24.85750 24.912521 24.963857 25.01173 25.056384 25.098024 25.136858
#> 2  19.67981 19.887551 20.081376 20.26214 20.430717 20.587935 20.734557
#> 3  17.20816 17.041383 16.885777 16.74066 16.605322 16.479106 16.361396
#> 4  19.88646 20.154521 20.404618 20.63786 20.855381 21.058242 21.247431
#> 5  18.23471 18.331500 18.421809 18.50603 18.584578 18.657831 18.726146
#> 6  18.29852 18.461759 18.614057 18.75609 18.888552 19.012086 19.127294
#> 7  14.43221 14.472144 14.509407 14.54416 14.576569 14.606794 14.634982
#> 8  16.45396 16.492943 16.529316 16.56324 16.594872 16.624375 16.651889
#> 9   7.55105  7.735391  7.907381  8.06778  8.217368  8.356874  8.486978
#> 10 28.99598 28.993136 28.990482 28.98801 28.985699 28.983546 28.981538
#> 11 27.84833 28.013250 28.167118 28.31062 28.444441 28.569248 28.685643
#> 12 24.15326 24.373163 24.578333 24.76968 24.948122 25.114541 25.269745
#> 13 17.61190 17.555785 17.503424 17.45459 17.409053 17.366582 17.326973
#> 14 18.15776 18.143481 18.130158 18.11773 18.106145 18.095339 18.085260
#> 15 15.57381 15.377759 15.194844 15.02426 14.865167 14.716800 14.578431
#> 16 15.71883 15.704720 15.691559 15.67929 15.667839 15.657164 15.647208
#> 17 25.71342 25.544712 25.387305 25.24051 25.103604 24.975926 24.856854
#> 18 19.41914 19.076801 18.757396 18.45952 18.181715 17.922636 17.681017
#> 19 21.87901 21.620227 21.378780 21.15361 20.943607 20.747762 20.565116
#> 20 17.07627 16.772182 16.488465 16.22387 15.977109 15.746978 15.532357
#> 21 22.14990 22.235072 22.314536 22.38864 22.457758 22.522214 22.582326
#>       0.9686    0.9033    0.8424    0.7856    0.7327    0.6833    0.6373
#> 1  25.173075 25.206851 25.238351 25.267728 25.295124 25.320675 25.344503
#> 2  20.871296 20.998820 21.117750 21.228663 21.332102 21.428569 21.518535
#> 3  16.251619 16.149241 16.053763 15.964720 15.881678 15.804233 15.732008
#> 4  21.423870 21.588417 21.741874 21.884988 22.018458 22.142931 22.259016
#> 5  18.789857 18.849275 18.904688 18.956366 19.004561 19.049508 19.091426
#> 6  19.234737 19.334940 19.428388 19.515539 19.596816 19.672615 19.743306
#> 7  14.661270 14.685787 14.708651 14.729975 14.749861 14.768407 14.785703
#> 8  16.677549 16.701480 16.723797 16.744611 16.764022 16.782125 16.799007
#> 9   8.608313  8.721471  8.827003  8.925422  9.017208  9.102807  9.182638
#> 10 28.979666 28.977920 28.976291 28.974772 28.973356 28.972035 28.970803
#> 11 28.794194 28.895428 28.989840 29.077889 29.160003 29.236583 29.308002
#> 12 25.414488 25.549476 25.675367 25.792772 25.902265 26.004379 26.099611
#> 13 17.290034 17.255584 17.223457 17.193494 17.165551 17.139491 17.115188
#> 14 18.075861 18.067095 18.058920 18.051296 18.044186 18.037555 18.031371
#> 15 14.449388 14.329043 14.216808 14.112137 14.014521 13.923484 13.838583
#> 16 15.637924 15.629265 15.621190 15.613658 15.606635 15.600085 15.593976
#> 17 24.745807 24.642244 24.545661 24.455587 24.371584 24.293243 24.220181
#> 18 17.455683 17.245536 17.049552 16.866777 16.696320 16.537351 16.389097
#> 19 20.394780 20.235924 20.087774 19.949609 19.820756 19.700587 19.588518
#> 20 15.332200 15.145534 14.971448 14.809095 14.657685 14.516478 14.384789
#> 21 22.638386 22.690669 22.739427 22.784900 22.827307 22.866857 22.903741
#>       0.5943    0.5542    0.5169    0.4821    0.4496    0.4193    0.3910
#> 1  25.355923 25.355930 25.355931 25.355931 25.355931 25.355931 25.431400
#> 2  21.561652 21.561680 21.561682 21.561682 21.561682 21.561682 21.533008
#> 3  15.697393 15.697371 15.697368 15.697368 15.697368 15.697368 15.600837
#> 4  22.314650 22.314686 22.314690 22.314690 22.314690 22.314690 22.299748
#> 5  19.111515 19.111528 19.111530 19.111530 19.111530 19.111530 19.073201
#> 6  19.777185 19.777207 19.777209 19.777209 19.777209 19.777209 19.756409
#> 7  14.793992 14.793997 14.793998 14.793998 14.793998 14.793998 14.795595
#> 8  16.807098 16.807104 16.807104 16.807104 16.807104 16.807104 16.778807
#> 9   9.220897  9.220922  9.220924  9.220925  9.220925  9.220925  9.363494
#> 10 28.970213 28.970212 28.970212 28.970212 28.970212 28.970212 28.968950
#> 11 29.342230 29.342253 29.342255 29.342255 29.342255 29.342255 29.404432
#> 12 26.145251 26.145280 26.145283 26.145283 26.145283 26.145283 26.054429
#> 13 17.103540 17.103532 17.103532 17.103532 17.103532 17.103532 17.031835
#> 14 18.028407 18.028405 18.028405 18.028405 18.028405 18.028405 17.960337
#> 15 13.797893 13.797867 13.797864 13.797864 13.797864 13.797864 13.717150
#> 16 15.591049 15.591047 15.591047 15.591046 15.591046 15.591046 15.558769
#> 17 24.185166 24.185143 24.185141 24.185141 24.185141 24.185141 24.185891
#> 18 16.318045 16.317998 16.317994 16.317994 16.317994 16.317994 16.362735
#> 19 19.534808 19.534773 19.534770 19.534769 19.534769 19.534769 19.566681
#> 20 14.321676 14.321635 14.321631 14.321631 14.321631 14.321631 14.411337
#> 21 22.921418 22.921429 22.921430 22.921430 22.921430 22.921430 23.044957
#>      0.3647   0.3401   0.3172   0.2958   0.2759   0.2573   0.2399   0.2238
#> 1  25.77394 26.07246 25.72394 25.72418 24.92672 24.92658 24.92646 24.92637
#> 2  21.40362 21.29088 20.74526 20.74522 19.76835 19.76806 19.76783 19.76764
#> 3  15.16238 14.78025 14.54401 14.54360 14.26176 14.26142 14.26114 14.26092
#> 4  22.23286 22.17459 22.00885 22.00891 22.26473 22.26490 22.26504 22.26515
#> 5  18.89965 18.74841 17.37033 17.37015 17.77028 17.77020 17.77013 17.77008
#> 6  19.66259 19.58085 18.60070 18.60063 19.14094 19.14101 19.14108 19.14113
#> 7  14.80297 14.80940 15.43798 15.43805 15.50792 15.50805 15.50815 15.50824
#> 8  16.65057 16.53882 17.41548 17.41548 17.36689 17.36698 17.36705 17.36710
#> 9  10.01086 10.57504 11.18760 11.18819 11.71116 11.71170 11.71213 11.71248
#> 10 28.96321 28.95822 28.78939 28.78938 29.60071 29.60092 29.60109 29.60122
#> 11 29.68704 29.93335 30.60504 30.60537 30.54938 30.54963 30.54984 30.55000
#> 12 25.64301 25.28449 25.37156 25.37138 25.29558 25.29556 25.29554 25.29552
#> 13 16.70640 16.42279 16.88663 16.88641 15.99911 15.99879 15.99853 15.99832
#> 14 17.65150 17.38236 18.16023 18.16007 18.04417 18.04414 18.04412 18.04410
#> 15 13.35035 13.03066 12.41756 12.41717 13.11567 13.11556 13.11548 13.11542
#> 16 15.41230 15.28465 15.75751 15.75743 15.65448 15.65446 15.65444 15.65443
#> 17 24.18873 24.19119 25.45453 25.45455 26.30885 26.30918 26.30945 26.30966
#> 18 16.56456 16.74041 16.19496 16.19490 16.65208 16.65202 16.65197 16.65194
#> 19 19.71059 19.83597 20.08954 20.08954 19.49029 19.49007 19.48988 19.48973
#> 20 14.81727 15.17100 14.55604 14.55614 13.98918 13.98890 13.98867 13.98849
#> 21 23.60561 24.09421 23.58286 23.58325 23.48175 23.48188 23.48197 23.48206
#>      0.2087   0.1946   0.1815   0.1693   0.1579   0.1472   0.1373   0.1280
#> 1  24.92630 24.92624 24.92619 24.92615 24.92612 24.92610 24.27228 24.27230
#> 2  19.76749 19.76737 19.76728 19.76720 19.76714 19.76709 19.90741 19.90759
#> 3  14.26074 14.26060 14.26049 14.26040 14.26032 14.26027 13.61535 13.61528
#> 4  22.26524 22.26531 22.26537 22.26542 22.26546 22.26549 22.31631 22.31606
#> 5  17.77005 17.77002 17.76999 17.76997 17.76996 17.76995 18.20192 18.20183
#> 6  19.14117 19.14121 19.14123 19.14126 19.14127 19.14129 20.01363 20.01364
#> 7  15.50830 15.50836 15.50840 15.50843 15.50846 15.50848 15.60272 15.60277
#> 8  17.36715 17.36718 17.36721 17.36723 17.36725 17.36726 17.87049 17.87066
#> 9  11.71276 11.71298 11.71316 11.71331 11.71342 11.71351 11.30322 11.30308
#> 10 29.60133 29.60142 29.60149 29.60154 29.60159 29.60163 29.58141 29.58138
#> 11 30.55013 30.55024 30.55032 30.55039 30.55044 30.55048 31.27091 31.27118
#> 12 25.29551 25.29550 25.29549 25.29548 25.29548 25.29547 24.73363 24.73344
#> 13 15.99814 15.99801 15.99790 15.99781 15.99774 15.99768 16.08230 16.08249
#> 14 18.04408 18.04406 18.04405 18.04404 18.04403 18.04403 18.56612 18.56631
#> 15 13.11536 13.11532 13.11529 13.11527 13.11524 13.11523 12.02160 12.02128
#> 16 15.65442 15.65441 15.65441 15.65440 15.65440 15.65439 15.66670 15.66677
#> 17 26.30984 26.30997 26.31008 26.31017 26.31024 26.31030 25.98130 25.98097
#> 18 16.65191 16.65188 16.65186 16.65185 16.65184 16.65183 17.46039 17.46055
#> 19 19.48961 19.48951 19.48944 19.48938 19.48933 19.48929 19.46517 19.46519
#> 20 13.98834 13.98823 13.98813 13.98806 13.98800 13.98795 14.46323 14.46344
#> 21 23.48212 23.48217 23.48221 23.48225 23.48227 23.48229 22.50394 22.50379
#>      0.1194   0.1114   0.1039   0.0969   0.0903   0.0842   0.0786   0.0733
#> 1  24.27232 24.27234 24.27236 24.27238 24.27240 23.81821 23.81811 23.81801
#> 2  19.90775 19.90789 19.90801 19.90811 19.90819 20.34460 20.34459 20.34458
#> 3  13.61521 13.61516 13.61512 13.61508 13.61505 13.99364 13.99392 13.99417
#> 4  22.31585 22.31566 22.31551 22.31537 22.31526 22.47195 22.47207 22.47218
#> 5  18.20176 18.20169 18.20163 18.20158 18.20154 18.13114 18.13095 18.13078
#> 6  20.01364 20.01363 20.01363 20.01363 20.01362 19.91682 19.91664 19.91648
#> 7  15.60282 15.60286 15.60290 15.60292 15.60295 14.45729 14.45702 14.45677
#> 8  17.87082 17.87094 17.87105 17.87115 17.87122 17.16588 17.16572 17.16557
#> 9  11.30295 11.30285 11.30276 11.30268 11.30262 12.12762 12.12788 12.12813
#> 10 29.58136 29.58135 29.58133 29.58132 29.58131 30.64158 30.64169 30.64180
#> 11 31.27142 31.27162 31.27178 31.27193 31.27205 31.19796 31.19788 31.19780
#> 12 24.73328 24.73315 24.73303 24.73294 24.73285 24.31815 24.31839 24.31861
#> 13 16.08265 16.08279 16.08291 16.08301 16.08309 16.38434 16.38427 16.38421
#> 14 18.56648 18.56662 18.56674 18.56684 18.56693 18.58551 18.58538 18.58525
#> 15 12.02100 12.02077 12.02058 12.02041 12.02027 11.70483 11.70492 11.70500
#> 16 15.66682 15.66686 15.66690 15.66693 15.66696 16.29779 16.29785 16.29790
#> 17 25.98069 25.98045 25.98024 25.98006 25.97992 25.84645 25.84655 25.84665
#> 18 17.46069 17.46081 17.46092 17.46100 17.46108 17.10847 17.10813 17.10781
#> 19 19.46520 19.46522 19.46523 19.46524 19.46525 19.99054 19.99073 19.99091
#> 20 14.46363 14.46378 14.46392 14.46403 14.46413 14.43605 14.43619 14.43632
#> 21 22.50365 22.50354 22.50345 22.50338 22.50331 21.96116 21.96111 21.96106
#>      0.0683   0.0637   0.0594   0.0554   0.0517   0.0482   0.0450   0.0419
#> 1  23.81792 23.81784 23.81776 23.81768 23.81761 23.81754 23.81748 23.81742
#> 2  20.34457 20.34455 20.34453 20.34451 20.34449 20.34447 20.34446 20.34444
#> 3  13.99441 13.99462 13.99481 13.99499 13.99515 13.99529 13.99543 13.99555
#> 4  22.47228 22.47237 22.47246 22.47254 22.47261 22.47268 22.47274 22.47280
#> 5  18.13063 18.13049 18.13036 18.13024 18.13014 18.13004 18.12995 18.12987
#> 6  19.91633 19.91619 19.91607 19.91596 19.91587 19.91578 19.91570 19.91563
#> 7  14.45655 14.45634 14.45616 14.45599 14.45583 14.45569 14.45557 14.45545
#> 8  17.16544 17.16532 17.16521 17.16511 17.16502 17.16494 17.16487 17.16481
#> 9  12.12835 12.12855 12.12874 12.12892 12.12908 12.12923 12.12937 12.12949
#> 10 30.64190 30.64199 30.64207 30.64215 30.64223 30.64230 30.64236 30.64242
#> 11 31.19773 31.19767 31.19762 31.19757 31.19752 31.19748 31.19745 31.19742
#> 12 24.31882 24.31901 24.31919 24.31935 24.31949 24.31963 24.31975 24.31987
#> 13 16.38414 16.38408 16.38402 16.38397 16.38391 16.38386 16.38382 16.38377
#> 14 18.58513 18.58502 18.58492 18.58483 18.58475 18.58467 18.58460 18.58454
#> 15 11.70508 11.70515 11.70521 11.70527 11.70533 11.70538 11.70542 11.70546
#> 16 16.29795 16.29799 16.29802 16.29805 16.29808 16.29810 16.29813 16.29814
#> 17 25.84674 25.84683 25.84691 25.84698 25.84705 25.84711 25.84717 25.84722
#> 18 17.10752 17.10725 17.10700 17.10677 17.10656 17.10637 17.10619 17.10603
#> 19 19.99108 19.99123 19.99137 19.99149 19.99161 19.99171 19.99181 19.99190
#> 20 14.43643 14.43654 14.43664 14.43672 14.43680 14.43688 14.43694 14.43700
#> 21 21.96101 21.96097 21.96093 21.96089 21.96086 21.96083 21.96080 21.96077
#>      0.0391   0.0365   0.0340   0.0317   0.0296   0.0276   0.0257   0.0240
#> 1  23.81737 23.81732 23.81727 23.81723 23.81718 23.81714 23.81711 23.81707
#> 2  20.34442 20.34440 20.34438 20.34437 20.34435 20.34434 20.34433 20.34432
#> 3  13.99566 13.99575 13.99584 13.99593 13.99600 13.99607 13.99613 13.99618
#> 4  22.47285 22.47290 22.47294 22.47298 22.47302 22.47305 22.47308 22.47311
#> 5  18.12980 18.12973 18.12967 18.12961 18.12956 18.12952 18.12947 18.12943
#> 6  19.91556 19.91550 19.91545 19.91540 19.91536 19.91532 19.91529 19.91526
#> 7  14.45534 14.45525 14.45516 14.45508 14.45501 14.45495 14.45489 14.45483
#> 8  17.16475 17.16469 17.16465 17.16460 17.16457 17.16453 17.16450 17.16447
#> 9  12.12961 12.12971 12.12981 12.12990 12.12998 12.13006 12.13013 12.13019
#> 10 30.64248 30.64253 30.64258 30.64262 30.64266 30.64270 30.64273 30.64277
#> 11 31.19739 31.19737 31.19735 31.19733 31.19732 31.19730 31.19729 31.19729
#> 12 24.31997 24.32006 24.32015 24.32023 24.32030 24.32037 24.32043 24.32048
#> 13 16.38373 16.38369 16.38366 16.38362 16.38359 16.38356 16.38354 16.38351
#> 14 18.58448 18.58443 18.58438 18.58434 18.58430 18.58427 18.58424 18.58421
#> 15 11.70550 11.70553 11.70556 11.70559 11.70561 11.70563 11.70565 11.70567
#> 16 16.29816 16.29818 16.29819 16.29820 16.29822 16.29823 16.29824 16.29824
#> 17 25.84727 25.84731 25.84735 25.84739 25.84743 25.84746 25.84749 25.84751
#> 18 17.10588 17.10575 17.10563 17.10551 17.10541 17.10532 17.10523 17.10516
#> 19 19.99198 19.99205 19.99211 19.99217 19.99223 19.99227 19.99232 19.99236
#> 20 14.43706 14.43711 14.43716 14.43720 14.43724 14.43728 14.43731 14.43735
#> 21 21.96074 21.96072 21.96069 21.96067 21.96065 21.96063 21.96061 21.96059
#>      0.0224   0.0209   0.0195   0.0181   0.0169   0.0158   0.0147   0.0137
#> 1  23.81704 23.81701 23.81698 23.81696 23.81694 23.81691 23.81689 23.81687
#> 2  20.34431 20.34430 20.34429 20.34428 20.34427 20.34427 20.34426 20.34426
#> 3  13.99623 13.99627 13.99631 13.99635 13.99638 13.99642 13.99644 13.99647
#> 4  22.47313 22.47316 22.47318 22.47320 22.47321 22.47323 22.47324 22.47326
#> 5  18.12940 18.12937 18.12934 18.12931 18.12929 18.12926 18.12924 18.12922
#> 6  19.91523 19.91521 19.91519 19.91517 19.91515 19.91513 19.91512 19.91511
#> 7  14.45478 14.45474 14.45469 14.45466 14.45462 14.45459 14.45456 14.45453
#> 8  17.16445 17.16443 17.16441 17.16439 17.16437 17.16436 17.16435 17.16434
#> 9  12.13025 12.13030 12.13035 12.13040 12.13044 12.13047 12.13051 12.13054
#> 10 30.64280 30.64282 30.64285 30.64287 30.64289 30.64291 30.64293 30.64295
#> 11 31.19728 31.19727 31.19727 31.19727 31.19727 31.19726 31.19726 31.19727
#> 12 24.32053 24.32058 24.32062 24.32066 24.32069 24.32072 24.32075 24.32078
#> 13 16.38349 16.38347 16.38345 16.38343 16.38341 16.38340 16.38338 16.38337
#> 14 18.58418 18.58416 18.58414 18.58412 18.58410 18.58409 18.58408 18.58406
#> 15 11.70568 11.70569 11.70570 11.70571 11.70572 11.70572 11.70573 11.70573
#> 16 16.29825 16.29826 16.29827 16.29827 16.29828 16.29829 16.29829 16.29830
#> 17 25.84753 25.84755 25.84757 25.84759 25.84761 25.84762 25.84763 25.84764
#> 18 17.10509 17.10503 17.10497 17.10492 17.10487 17.10483 17.10480 17.10476
#> 19 19.99239 19.99242 19.99245 19.99248 19.99250 19.99252 19.99254 19.99256
#> 20 14.43737 14.43740 14.43743 14.43745 14.43747 14.43749 14.43751 14.43753
#> 21 21.96058 21.96056 21.96055 21.96053 21.96052 21.96050 21.96049 21.96048
#>      0.0128   0.0119   0.0111   0.0104   0.0097   0.0090   0.0084   0.0079
#> 1  23.81685 23.81684 23.81682 23.81681 23.81679 23.81678 23.81677 23.81676
#> 2  20.34425 20.34425 20.34425 20.34424 20.34424 20.34424 20.34424 20.34424
#> 3  13.99649 13.99651 13.99653 13.99655 13.99656 13.99658 13.99659 13.99660
#> 4  22.47327 22.47328 22.47329 22.47330 22.47331 22.47331 22.47332 22.47333
#> 5  18.12920 18.12919 18.12917 18.12916 18.12914 18.12913 18.12912 18.12911
#> 6  19.91509 19.91508 19.91507 19.91507 19.91506 19.91505 19.91504 19.91504
#> 7  14.45451 14.45449 14.45447 14.45445 14.45443 14.45441 14.45440 14.45438
#> 8  17.16433 17.16432 17.16431 17.16430 17.16429 17.16429 17.16428 17.16428
#> 9  12.13057 12.13059 12.13062 12.13064 12.13066 12.13068 12.13069 12.13071
#> 10 30.64296 30.64297 30.64299 30.64300 30.64301 30.64302 30.64302 30.64303
#> 11 31.19727 31.19727 31.19727 31.19727 31.19727 31.19728 31.19728 31.19728
#> 12 24.32080 24.32082 24.32084 24.32086 24.32088 24.32089 24.32091 24.32092
#> 13 16.38336 16.38334 16.38333 16.38332 16.38332 16.38331 16.38330 16.38329
#> 14 18.58405 18.58404 18.58403 18.58403 18.58402 18.58401 18.58401 18.58400
#> 15 11.70573 11.70574 11.70574 11.70574 11.70574 11.70574 11.70574 11.70574
#> 16 16.29830 16.29831 16.29831 16.29831 16.29832 16.29832 16.29833 16.29833
#> 17 25.84765 25.84766 25.84767 25.84767 25.84768 25.84768 25.84769 25.84769
#> 18 17.10473 17.10470 17.10468 17.10466 17.10464 17.10462 17.10461 17.10459
#> 19 19.99257 19.99258 19.99260 19.99261 19.99262 19.99262 19.99263 19.99264
#> 20 14.43754 14.43756 14.43757 14.43759 14.43760 14.43761 14.43762 14.43764
#> 21 21.96047 21.96045 21.96044 21.96043 21.96042 21.96041 21.96040 21.96040
#>      0.0073   0.0068   0.0064   0.0059   0.0055   0.0052   0.0048
#> 1  23.81675 23.81674 23.81673 23.81673 23.81672 23.81671 23.81671
#> 2  20.34424 20.34424 20.34424 20.34424 20.34424 20.34424 20.34424
#> 3  13.99661 13.99663 13.99663 13.99664 13.99665 13.99666 13.99667
#> 4  22.47333 22.47334 22.47334 22.47335 22.47335 22.47336 22.47336
#> 5  18.12910 18.12909 18.12908 18.12907 18.12906 18.12906 18.12905
#> 6  19.91503 19.91503 19.91502 19.91502 19.91501 19.91501 19.91501
#> 7  14.45437 14.45436 14.45434 14.45433 14.45432 14.45432 14.45431
#> 8  17.16427 17.16427 17.16427 17.16426 17.16426 17.16426 17.16426
#> 9  12.13072 12.13073 12.13075 12.13076 12.13077 12.13077 12.13078
#> 10 30.64304 30.64304 30.64305 30.64305 30.64306 30.64306 30.64307
#> 11 31.19729 31.19729 31.19729 31.19729 31.19730 31.19730 31.19730
#> 12 24.32093 24.32095 24.32096 24.32097 24.32098 24.32098 24.32099
#> 13 16.38329 16.38328 16.38328 16.38327 16.38327 16.38326 16.38326
#> 14 18.58400 18.58399 18.58399 18.58399 18.58398 18.58398 18.58398
#> 15 11.70574 11.70574 11.70574 11.70573 11.70573 11.70573 11.70573
#> 16 16.29833 16.29834 16.29834 16.29834 16.29835 16.29835 16.29835
#> 17 25.84770 25.84770 25.84770 25.84770 25.84771 25.84771 25.84771
#> 18 17.10458 17.10457 17.10456 17.10455 17.10454 17.10454 17.10453
#> 19 19.99264 19.99265 19.99265 19.99265 19.99266 19.99266 19.99266
#> 20 14.43765 14.43766 14.43766 14.43767 14.43768 14.43769 14.43770
#> 21 21.96039 21.96038 21.96037 21.96037 21.96036 21.96035 21.96035
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 22.8 21.4 14.3 24.4 19.2 17.8 16.4 17.3 10.4 30.4 33.9 21.5 15.5 15.2 13.3
#> [16] 19.2 26.0 15.8 19.7 15.0 21.4
#> 
#> $X
#>               am        carb        cyl       disp        drat       gear
#>  [1,]  1.2747549 -1.12763391 -1.2865350 -1.0924519  0.39895570  0.3735437
#>  [2,] -0.7844645 -1.12763391 -0.1608169  0.2106963 -0.99738924 -0.9338592
#>  [3,] -0.7844645  0.60507186  0.9649013  1.0968371 -0.76164270 -0.9338592
#>  [4,] -0.7844645 -0.55006532 -1.2865350 -0.7562397  0.10880610  0.3735437
#>  [5,] -0.7844645  0.60507186 -0.1608169 -0.5746677  0.52589615  0.3735437
#>  [6,] -0.7844645  0.60507186 -0.1608169 -0.5746677  0.52589615  0.3735437
#>  [7,] -0.7844645  0.02750327  0.9649013  0.3653366 -1.01552359 -0.9338592
#>  [8,] -0.7844645  0.02750327  0.9649013  0.3653366 -1.01552359 -0.9338592
#>  [9,] -0.7844645  0.60507186  0.9649013  2.0698544 -1.26940449 -0.9338592
#> [10,]  1.2747549 -0.55006532 -1.2865350 -1.3730632  2.35746549  0.3735437
#> [11,]  1.2747549 -1.12763391 -1.2865350 -1.4130264  1.06992664  0.3735437
#> [12,] -0.7844645 -1.12763391 -1.2865350 -0.9873313  0.12694045 -0.9338592
#> [13,] -0.7844645 -0.55006532  0.9649013  0.7319556 -1.57768844 -0.9338592
#> [14,] -0.7844645 -0.55006532  0.9649013  0.6103284 -0.87044880 -0.9338592
#> [15,] -0.7844645  0.60507186  0.9649013  1.0099606  0.18134350 -0.9338592
#> [16,] -0.7844645 -0.55006532  0.9649013  1.4443433 -0.99738924 -0.9338592
#> [17,]  1.2747549 -0.55006532 -1.2865350 -0.9855938  1.45074799  1.6809466
#> [18,]  1.2747549  0.60507186  0.9649013  1.0186482  1.06992664  1.6809466
#> [19,]  1.2747549  1.76020904 -0.1608169 -0.7710087 -0.01813435  1.6809466
#> [20,]  1.2747549  2.91534622  0.9649013  0.5842655 -0.16320915  1.6809466
#> [21,]  1.2747549 -0.55006532 -1.2865350 -0.9795124  0.87044880  0.3735437
#>                hp        qsec         vs          wt
#>  [1,] -0.83992896  0.62348988  1.1547005 -1.08658796
#>  [2,] -0.60493316  1.12245355  1.1547005  0.01792371
#>  [3,]  1.26120995 -1.04172621 -0.8660254  0.45602610
#>  [4,] -1.26845071  1.45910373  1.1547005 -0.01292858
#>  [5,] -0.42523049  0.43712996  1.1547005  0.29559424
#>  [6,] -0.42523049  0.79782658  1.1547005  0.29559424
#>  [7,]  0.36269660 -0.10391498 -0.8660254  1.07307173
#>  [8,]  0.36269660  0.01631723 -0.8660254  0.65348070
#>  [9,]  0.70827866  0.24475842 -0.8660254  2.52929940
#> [10,] -1.40668354  0.56938539  1.1547005 -1.95662230
#> [11,] -1.22698087  1.39898763  1.1547005 -1.68512222
#> [12,] -0.78463583  1.46511534  1.1547005 -0.90764473
#> [13,] -0.05200187 -0.42253033 -0.8660254  0.39432154
#> [14,] -0.05200187 -0.16403108 -0.8660254  0.28942378
#> [15,]  1.26120995 -1.30022546 -0.8660254  0.78923074
#> [16,]  0.29358019 -0.31432135 -0.8660254  0.79540119
#> [17,] -0.86757553 -0.52472771 -0.8660254 -1.30872439
#> [18,]  1.52385231 -1.84728201 -0.8660254 -0.03761040
#> [19,]  0.29358019 -1.24612096 -0.8660254 -0.53124690
#> [20,]  2.50530536 -1.78716590 -0.8660254  0.45602610
#> [21,] -0.61875645  0.61747827  1.1547005 -0.51890599
#> attr(,"center")
#>  [1]   0.3809524   2.9523810   6.2857143 233.7476190   3.6300000   3.7142857
#>  [7] 153.7619048  17.5728571   0.4285714   3.2004762
#> attr(,"scale")
#>  [1]   0.4856209   1.7313961   1.7766436 115.1058627   0.5514397   0.7648752
#>  [7]  72.3417187   1.6634478   0.4948717   0.8103129
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 33
#> 
#> $lambda.min
#> [1] 0.5168946
#> 
#> $null.dev
#> [1] 30.89175
#> 
#> $Bias
#> [1] 3.438427
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 9.843653