Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 39.13454 35.79659 31.19945 27.46260 24.45636 22.06912 20.20494 18.78138
#>   [9] 17.72773 16.98337 16.49642 16.22254 16.12388 16.16823 16.32701 16.54123
#>  [17] 16.29168 16.10943 16.10992 16.02036 15.85958 15.63810 15.22090 14.50844
#>  [25] 14.82081 14.27555 13.96780 13.54481 13.29740 13.08713 13.56156 13.41925
#>  [33] 15.82005 15.74715 15.63457 15.56875 15.55256 15.07411 15.14641 15.32561
#>  [41] 14.29567 14.76043 14.74219 14.73340 14.73354 15.51040 15.44421 15.60574
#>  [49] 20.91193 22.05220 28.79667 29.89593 30.02048 32.49496 32.66190 32.79367
#>  [57] 54.22795 54.98735 56.16170 56.32821 56.33028 56.33223 60.53059 60.51729
#>  [65] 60.51999 59.90206 62.60437 62.21454 62.21758 62.22064 62.22372 62.15852
#>  [73] 62.16166 62.16479 62.16791 62.15576 62.09716 62.06572 62.52207 62.52691
#>  [81] 62.53161 62.53620 62.54066 62.54501 62.54925 62.55337 62.55739 62.56131
#>  [89] 62.23361 62.23676 62.23984 62.24287 62.24583 62.24874 62.25158 62.25437
#>  [97] 62.25710 62.25977 62.26239 62.26496
#> 
#> $cvse
#>   [1] 12.006626 11.718753 10.618242  9.629846  8.734902  7.920873  7.180258
#>   [8]  6.509722  5.909410  5.382343  4.933723  4.569908  4.296837  4.117966
#>  [15]  4.032538  4.035465  3.943971  3.884030  3.891696  3.892340  3.887915
#>  [22]  3.883158  3.881828  3.746632  3.538934  3.482938  3.416651  3.366817
#>  [29]  3.370431  3.379280  3.369147  3.374648  4.447412  4.444917  4.441513
#>  [36]  4.438463  4.435899  4.441419  4.441791  4.472019  4.087363  4.024916
#>  [43]  4.022010  4.020546  4.020483  4.243173  4.103485  4.063906  6.282157
#>  [50]  6.407108  9.557905 10.176695 10.167151 10.128730 10.099176 10.084960
#>  [57] 25.421715 25.385189 25.429082 25.418627 25.419209 25.419737 26.758787
#>  [64] 26.759294 26.760237 26.749767 28.629797 28.633187 28.636106 28.638972
#>  [71] 28.641786 28.647589 28.650302 28.652966 28.655581 28.663897 28.670551
#>  [78] 28.675603 28.683534 28.685897 28.688217 28.690496 28.692733 28.694929
#>  [85] 28.697085 28.699202 28.701279 28.703317 28.707802 28.709784 28.711729
#>  [92] 28.713636 28.715507 28.717342 28.719141 28.720904 28.722634 28.724330
#>  [99] 28.725992 28.727621
#> 
#> $fold
#>  [1]  8  1  7  9  4  6  2  5  1  3  4  9  6  1  3 10 10  7  8  2  5
#> 
#> $lambda
#>   [1] 5.128440121 4.782800421 4.460455680 4.159835896 3.879476879 3.618013122
#>   [7] 3.374171146 3.146763304 2.934681989 2.736894245 2.552436733 2.380411040
#>  [13] 2.219979303 2.070360128 1.930824784 1.800693656 1.679332931 1.566151512
#>  [19] 1.460598141 1.362158715 1.270353777 1.184736184 1.104888930 1.030423114
#>  [25] 0.960976045 0.896209476 0.835807957 0.779477298 0.726943137 0.677949602
#>  [31] 0.632258068 0.589645990 0.549905824 0.512844012 0.478280042 0.446045568
#>  [37] 0.415983590 0.387947688 0.361801312 0.337417115 0.314676330 0.293468199
#>  [43] 0.273689425 0.255243673 0.238041104 0.221997930 0.207036013 0.193082478
#>  [49] 0.180069366 0.167933293 0.156615151 0.146059814 0.136215872 0.127035378
#>  [55] 0.118473618 0.110488893 0.103042311 0.096097605 0.089620948 0.083580796
#>  [61] 0.077947730 0.072694314 0.067794960 0.063225807 0.058964599 0.054990582
#>  [67] 0.051284401 0.047828004 0.044604557 0.041598359 0.038794769 0.036180131
#>  [73] 0.033741711 0.031467633 0.029346820 0.027368942 0.025524367 0.023804110
#>  [79] 0.022199793 0.020703601 0.019308248 0.018006937 0.016793329 0.015661515
#>  [85] 0.014605981 0.013621587 0.012703538 0.011847362 0.011048889 0.010304231
#>  [91] 0.009609760 0.008962095 0.008358080 0.007794773 0.007269431 0.006779496
#>  [97] 0.006322581 0.005896460 0.005499058 0.005128440
#> 
#> $fit
#> $beta
#>               5.1284     4.7828     4.4605     4.1598    3.8795    3.6180
#> (Intercept) 20.67619 22.5487258 24.2950586 25.9236944 27.442566 28.859070
#> am           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> cyl          0.00000 -0.3072128 -0.5937205 -0.8609186 -1.110108 -1.342504
#> disp         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> wt           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#>                3.3742    3.1468    2.9347    2.7369    2.5524    2.3804
#> (Intercept) 30.180107 31.412110 32.561080 33.632614 34.631929 35.563895
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.559236 -1.761362 -1.949865 -2.125663 -2.289613 -2.442514
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>               2.2200    2.0704    1.9308    1.8007    1.6793    1.5662
#> (Intercept) 36.43305 37.243625 37.999570 38.704568 39.198726 39.198726
#> am           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.58511 -2.718095 -2.842117 -2.957781 -3.038854 -3.038854
#> disp         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.4606    1.3622    1.2704    1.1847     1.1049    1.0304
#> (Intercept) 39.198726 39.198726 39.198726 39.198726 39.6856645 40.144827
#> am           0.000000  0.000000  0.000000  0.000000  0.0000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.0000000  0.000000
#> cyl         -3.038854 -3.038854 -3.038854 -3.038854 -2.6909326 -2.361806
#> disp         0.000000  0.000000  0.000000  0.000000  0.0000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.0000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.0000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.0000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.0000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.0000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000 -0.8172218 -1.589838
#>                0.9610    0.8962       0.8358      0.7795     0.7269     0.6779
#> (Intercept) 40.573119 40.894554 40.895423939 40.92089764 40.9446437 40.9672635
#> am           0.000000  0.000000  0.000000000  0.00000000  0.0000000  0.0000000
#> carb         0.000000  0.000000 -0.005787357 -0.09551108 -0.1788983 -0.2566351
#> cyl         -2.054577 -1.823652 -1.821298173 -1.79125306 -1.7633682 -1.7376330
#> disp         0.000000  0.000000  0.000000000  0.00000000  0.0000000  0.0000000
#> drat         0.000000  0.000000  0.000000000  0.00000000  0.0000000  0.0000000
#> gear         0.000000  0.000000  0.000000000  0.00000000  0.0000000  0.0000000
#> hp           0.000000  0.000000  0.000000000  0.00000000  0.0000000  0.0000000
#> qsec         0.000000  0.000000  0.000000000  0.00000000  0.0000000  0.0000000
#> vs           0.000000  0.000000  0.000000000  0.00000000  0.0000000  0.0000000
#> wt          -2.310946 -2.852808 -2.852741424 -2.84313338 -2.8341528 -2.8254354
#>                 0.6323     0.5896     0.5499     0.5128     0.4783     0.4460
#> (Intercept) 40.9883791 41.0080723 41.0264384 41.0435666 41.0595404 41.0744377
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb        -0.3291145 -0.3967082 -0.4597463 -0.5185358 -0.5733631 -0.6244953
#> cyl         -1.7136518 -1.6912879 -1.6704312 -1.6509802 -1.6328401 -1.6159226
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.8172894 -2.8096917 -2.8026060 -2.7959979 -2.7898351 -2.7840877
#>                 0.4160     0.3879     0.3618    0.3374    0.3147    0.2935
#> (Intercept) 41.0883309 41.1012878 41.1133714 37.722328 37.722092 37.721956
#> am           0.0000000  0.0000000  0.0000000  2.206496  2.206870  2.207111
#> carb        -0.6721813 -0.7166534 -0.7581282 -1.095094 -1.095103 -1.095106
#> cyl         -1.6001453 -1.5854313 -1.5717090 -1.528165 -1.528363 -1.528501
#> disp         0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> drat         0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> gear         0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> hp           0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> qsec         0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> vs           0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> wt          -2.7787276 -2.7737288 -2.7690669 -1.804244 -1.803834 -1.803558
#>                0.2737    0.2552    0.2380    0.2220    0.2070    0.1931
#> (Intercept) 37.721881 37.721842 37.721825 37.721820 37.721821 37.721825
#> am           2.207264  2.207359  2.207418  2.207453  2.207474  2.207485
#> carb        -1.095104 -1.095101 -1.095098 -1.095095 -1.095092 -1.095090
#> cyl         -1.528596 -1.528662 -1.528706 -1.528736 -1.528755 -1.528769
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -1.803374 -1.803253 -1.803174 -1.803123 -1.803090 -1.803070
#>                0.1801    0.1679    0.1566    0.1461    0.1362    0.1270
#> (Intercept) 37.721829 37.721834 37.721837 37.721841 37.721843 37.721845
#> am           2.207491  2.207494  2.207495  2.207495  2.207494  2.207493
#> carb        -1.095088 -1.095086 -1.095085 -1.095085 -1.095084 -1.095084
#> cyl         -1.528777 -1.528782 -1.528786 -1.528788 -1.528789 -1.528790
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -1.803057 -1.803050 -1.803045 -1.803043 -1.803042 -1.803041
#>                 0.1185     0.1105     0.1030      0.0961      0.0896
#> (Intercept) 38.7186553 38.7203453 38.7218060 31.23603612 31.23437270
#> am           2.3543922  2.3547281  2.3550660  2.97706744  2.97674715
#> carb        -0.9080517 -0.9081515 -0.9082065  0.00000000  0.00000000
#> cyl         -1.6056659 -1.6055891 -1.6055684 -1.93754948 -1.93725594
#> disp         0.0000000  0.0000000  0.0000000  0.03099550  0.03099124
#> drat         1.0089779  1.0081529  1.0075562  1.15037349  1.15061928
#> gear        -1.0841771 -1.0838994 -1.0837615 -1.23122410 -1.23105064
#> hp           0.0000000  0.0000000  0.0000000 -0.01953761 -0.01954223
#> qsec         0.0000000  0.0000000  0.0000000  0.56710402  0.56705324
#> vs           0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> wt          -2.0073339 -2.0073666 -2.0073511 -4.41165649 -4.41134058
#>                  0.0836      0.0779      0.0727      0.0678      0.0632
#> (Intercept) 31.23245176 31.56936465 31.57150897 31.57361516 31.57568305
#> am           2.97646021  3.24705746  3.24709232  3.24712488  3.24715525
#> carb         0.00000000  0.65373794  0.65396270  0.65418059  0.65439180
#> cyl         -1.93695986 -2.09666008 -2.09698209 -2.09729672 -2.09760403
#> disp         0.03098749  0.04374072  0.04374548  0.04375009  0.04375453
#> drat         1.15085246  1.28250096  1.28239644  1.28229420  1.28219421
#> gear        -1.23085998 -1.75162907 -1.75200708 -1.75237604 -1.75273607
#> hp          -0.01954701 -0.02818089 -0.02817863 -0.02817638 -0.02817415
#> qsec         0.56701483  0.66177257  0.66185233  0.66192956  0.66200432
#> vs           0.00000000  0.93048686  0.93035758  0.93022942  0.93010245
#> wt          -4.41105785 -5.52308098 -5.52363931 -5.52418091 -5.52470617
#>                  0.0590      0.0550      0.0513      0.0478      0.0446
#> (Intercept) 31.57771257 31.57970365 31.58165630 31.58357056 31.58544651
#> am           3.24718352  3.24720979  3.24723415  3.24725668  3.24727746
#> carb         0.65459646  0.65479474  0.65498680  0.65517277  0.65535282
#> cyl         -2.09790412 -2.09819708 -2.09848298 -2.09876192 -2.09903401
#> disp         0.04375883  0.04376298  0.04376699  0.04377086  0.04377460
#> drat         1.28209644  1.28200089  1.28190753  1.28181633  1.28172727
#> gear        -1.75308727 -1.75342978 -1.75376372 -1.75408923 -1.75440641
#> hp          -0.02817193 -0.02816973 -0.02816756 -0.02816540 -0.02816327
#> qsec         0.66207667  0.66214668  0.66221440  0.66227989  0.66234322
#> vs           0.92997676  0.92985241  0.92972946  0.92960798  0.92948802
#> wt          -5.52521545 -5.52570912 -5.52618756 -5.52665111 -5.52710014
#>                  0.0416      0.0388      0.0362      0.0337      0.0315
#> (Intercept) 31.58728430 31.58908406 31.59084601 31.59257037 31.59425740
#> am           3.24729659  3.24731413  3.24733016  3.24734475  3.24735797
#> carb         0.65552709  0.65569572  0.65585886  0.65601665  0.65616922
#> cyl         -2.09929932 -2.09955797 -2.09981005 -2.10005567 -2.10029492
#> disp         0.04377820  0.04378168  0.04378503  0.04378827  0.04379138
#> drat         1.28164033  1.28155548  1.28147269  1.28139194  1.28131319
#> gear        -1.75471542 -1.75501638 -1.75530943 -1.75559470 -1.75587233
#> hp          -0.02816116 -0.02815907 -0.02815701 -0.02815498 -0.02815298
#> qsec         0.66240443  0.66246359  0.66252074  0.66257595  0.66262926
#> vs           0.92936962  0.92925283  0.92913768  0.92902422  0.92891246
#> wt          -5.52753499 -5.52795603 -5.52836358 -5.52875798 -5.52913958
#>                  0.0293      0.0274      0.0255      0.0238      0.0222
#> (Intercept) 31.59590739 31.59752064 31.59909749 31.60063831 31.60214347
#> am           3.24736989  3.24738058  3.24739009  3.24739848  3.24740582
#> carb         0.65631672  0.65645928  0.65659702  0.65673009  0.65685860
#> cyl         -2.10052792 -2.10075477 -2.10097558 -2.10119044 -2.10139948
#> disp         0.04379439  0.04379728  0.04380007  0.04380275  0.04380534
#> drat         1.28123641  1.28116159  1.28108868  1.28101766  1.28094849
#> gear        -1.75614246 -1.75640522 -1.75666075 -1.75690919 -1.75715068
#> hp          -0.02815100 -0.02814905 -0.02814713 -0.02814524 -0.02814338
#> qsec         0.66268073  0.66273041  0.66277834  0.66282459  0.66286919
#> vs           0.92880245  0.92869420  0.92858773  0.92848306  0.92838020
#> wt          -5.52950868 -5.52986563 -5.53021073 -5.53054430 -5.53086663
#>                  0.0207      0.0193      0.0180      0.0168      0.0157
#> (Intercept) 31.60361336 31.60504841 31.60644903 31.60781568 31.60914881
#> am           3.24741215  3.24741754  3.24742202  3.24742565  3.24742848
#> carb         0.65698268  0.65710245  0.65721803  0.65732955  0.65743711
#> cyl         -2.10160280 -2.10180050 -2.10199270 -2.10217950 -2.10236100
#> disp         0.04380782  0.04381021  0.04381251  0.04381473  0.04381685
#> drat         1.28088115  1.28081559  1.28075180  1.28068974  1.28062938
#> gear        -1.75738535 -1.75761335 -1.75783481 -1.75804986 -1.75825864
#> hp          -0.02814155 -0.02813976 -0.02813799 -0.02813625 -0.02813455
#> qsec         0.66291219  0.66295364  0.66299359  0.66303208  0.66306914
#> vs           0.92827916  0.92817996  0.92808258  0.92798705  0.92789335
#> wt          -5.53117804 -5.53147882 -5.53176926 -5.53204964 -5.53232024
#>                  0.0146      0.0136      0.0127      0.0118      0.0110
#> (Intercept) 31.61044889 31.61171640 31.61295181 31.61415564 31.61532838
#> am           3.24743056  3.24743191  3.24743260  3.24743264  3.24743210
#> carb         0.65754082  0.65764081  0.65773717  0.65783002  0.65791945
#> cyl         -2.10253733 -2.10270858 -2.10287486 -2.10303628 -2.10319294
#> disp         0.04381890  0.04382086  0.04382274  0.04382455  0.04382629
#> drat         1.28057068  1.28051361  1.28045815  1.28040425  1.28035189
#> gear        -1.75846129 -1.75865794 -1.75884872 -1.75903376 -1.75921320
#> hp          -0.02813287 -0.02813123 -0.02812962 -0.02812804 -0.02812649
#> qsec         0.66310483  0.66313918  0.66317224  0.66320405  0.66323464
#> vs           0.92780149  0.92771146  0.92762326  0.92753688  0.92745231
#> wt          -5.53258134 -5.53283320 -5.53307609 -5.53331026 -5.53353597
#>                  0.0103      0.0096      0.0090      0.0084      0.0078
#> (Intercept) 31.61647053 31.61758262 31.61866516 31.61971866 31.62074366
#> am           3.24743099  3.24742936  3.24742724  3.24742465  3.24742164
#> carb         0.65800557  0.65808848  0.65816827  0.65824504  0.65831888
#> cyl         -2.10334494 -2.10349239 -2.10363540 -2.10377406 -2.10390847
#> disp         0.04382795  0.04382955  0.04383108  0.04383255  0.04383395
#> drat         1.28030104  1.28025166  1.28020372  1.28015720  1.28011206
#> gear        -1.75938716 -1.75955576 -1.75971915 -1.75987743 -1.76003073
#> hp          -0.02812497 -0.02812349 -0.02812203 -0.02812061 -0.02811922
#> qsec         0.66326404  0.66329231  0.66331947  0.66334556  0.66337061
#> vs           0.92736953  0.92728855  0.92720934  0.92713188  0.92705617
#> wt          -5.53375347 -5.53396298 -5.53416476 -5.53435904 -5.53454602
#>                  0.0073      0.0068      0.0063      0.0059      0.0055
#> (Intercept) 31.62174067 31.62271022 31.62365283 31.62456904 31.62545936
#> am           3.24741823  3.24741445  3.24741032  3.24740588  3.24740114
#> carb         0.65838989  0.65845814  0.65852373  0.65858674  0.65864725
#> cyl         -2.10403873 -2.10416494 -2.10428721 -2.10440561 -2.10452025
#> disp         0.04383530  0.04383658  0.04383782  0.04383900  0.04384012
#> drat         1.28006827  1.28002579  1.27998461  1.27994469  1.27990600
#> gear        -1.76017918 -1.76032289 -1.76046199 -1.76059658 -1.76072678
#> hp          -0.02811785 -0.02811652 -0.02811522 -0.02811394 -0.02811270
#> qsec         0.66339466  0.66341773  0.66343987  0.66346110  0.66348144
#> vs           0.92698219  0.92690992  0.92683933  0.92677042  0.92670315
#> wt          -5.53472595 -5.53489904 -5.53506548 -5.53522550 -5.53537929
#>                  0.0051
#> (Intercept) 31.62632433
#> am           3.24739612
#> carb         0.65870534
#> cyl         -2.10463123
#> disp         0.04384120
#> drat         1.27986851
#> gear        -1.76085271
#> hp          -0.02811149
#> qsec         0.66350094
#> vs           0.92663752
#> wt          -5.53552705
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1  34  33  33  16   5   6   5   5   5   5   5   5   5   5
#>  [37]   5   5   5  57   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [55]  54   1   1 242   1   1 289   1   1   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.128440121 4.782800421 4.460455680 4.159835896 3.879476879 3.618013122
#>   [7] 3.374171146 3.146763304 2.934681989 2.736894245 2.552436733 2.380411040
#>  [13] 2.219979303 2.070360128 1.930824784 1.800693656 1.679332931 1.566151512
#>  [19] 1.460598141 1.362158715 1.270353777 1.184736184 1.104888930 1.030423114
#>  [25] 0.960976045 0.896209476 0.835807957 0.779477298 0.726943137 0.677949602
#>  [31] 0.632258068 0.589645990 0.549905824 0.512844012 0.478280042 0.446045568
#>  [37] 0.415983590 0.387947688 0.361801312 0.337417115 0.314676330 0.293468199
#>  [43] 0.273689425 0.255243673 0.238041104 0.221997930 0.207036013 0.193082478
#>  [49] 0.180069366 0.167933293 0.156615151 0.146059814 0.136215872 0.127035378
#>  [55] 0.118473618 0.110488893 0.103042311 0.096097605 0.089620948 0.083580796
#>  [61] 0.077947730 0.072694314 0.067794960 0.063225807 0.058964599 0.054990582
#>  [67] 0.051284401 0.047828004 0.044604557 0.041598359 0.038794769 0.036180131
#>  [73] 0.033741711 0.031467633 0.029346820 0.027368942 0.025524367 0.023804110
#>  [79] 0.022199793 0.020703601 0.019308248 0.018006937 0.016793329 0.015661515
#>  [85] 0.014605981 0.013621587 0.012703538 0.011847362 0.011048889 0.010304231
#>  [91] 0.009609760 0.008962095 0.008358080 0.007794773 0.007269431 0.006779496
#>  [97] 0.006322581 0.005896460 0.005499058 0.005128440
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 22
#> 
#> $loss
#>   [1] 760.25810 654.22957 565.52095 491.63976 430.43411 380.04729 338.87827
#>   [8] 305.54749 278.86715 257.81539 241.51395 229.20875 220.25303 214.09281
#>  [15] 210.25418 208.33235 207.93924 207.93924 207.93924 207.93924 207.93924
#>  [22] 207.93924 174.66258 153.43184 142.58690 140.13703 139.85415 135.74251
#>  [29] 132.38279 129.65118 127.45242 125.70459 124.33784 123.29221 122.51621
#>  [36] 121.96573 121.60302 121.39580 121.31651 113.25138 113.25138 113.25138
#>  [43] 113.25137 113.25137 113.25137 113.25137 113.25137 113.25137 113.25137
#>  [50] 113.25137 113.25137 113.25137 113.25137 113.25137 110.14860 110.14860
#>  [57] 110.14859  97.51783  97.51781  97.51780  95.34012  95.34009  95.34007
#>  [64]  95.34005  95.34003  95.34001  95.33999  95.33997  95.33996  95.33994
#>  [71]  95.33993  95.33992  95.33990  95.33989  95.33988  95.33987  95.33986
#>  [78]  95.33985  95.33984  95.33984  95.33983  95.33982  95.33981  95.33981
#>  [85]  95.33980  95.33980  95.33979  95.33979  95.33978  95.33978  95.33978
#>  [92]  95.33977  95.33977  95.33977  95.33976  95.33976  95.33976  95.33975
#>  [99]  95.33975  95.33975
#> 
#> $linear.predictors
#>      5.1284   4.7828   4.4605   4.1598   3.8795   3.6180   3.3742   3.1468
#> 1  20.67619 20.70545 20.73274 20.75818 20.78192 20.80405 20.82469 20.84394
#> 2  20.67619 20.70545 20.73274 20.75818 20.78192 20.80405 20.82469 20.84394
#> 3  20.67619 21.31987 21.92018 22.48002 23.00213 23.48906 23.94316 24.36666
#> 4  20.67619 20.70545 20.73274 20.75818 20.78192 20.80405 20.82469 20.84394
#> 5  20.67619 20.09102 19.54529 19.03635 18.56170 18.11904 17.70622 17.32122
#> 6  20.67619 20.70545 20.73274 20.75818 20.78192 20.80405 20.82469 20.84394
#> 7  20.67619 21.31987 21.92018 22.48002 23.00213 23.48906 23.94316 24.36666
#> 8  20.67619 21.31987 21.92018 22.48002 23.00213 23.48906 23.94316 24.36666
#> 9  20.67619 20.70545 20.73274 20.75818 20.78192 20.80405 20.82469 20.84394
#> 10 20.67619 20.09102 19.54529 19.03635 18.56170 18.11904 17.70622 17.32122
#> 11 20.67619 20.09102 19.54529 19.03635 18.56170 18.11904 17.70622 17.32122
#> 12 20.67619 20.09102 19.54529 19.03635 18.56170 18.11904 17.70622 17.32122
#> 13 20.67619 21.31987 21.92018 22.48002 23.00213 23.48906 23.94316 24.36666
#> 14 20.67619 21.31987 21.92018 22.48002 23.00213 23.48906 23.94316 24.36666
#> 15 20.67619 20.09102 19.54529 19.03635 18.56170 18.11904 17.70622 17.32122
#> 16 20.67619 20.09102 19.54529 19.03635 18.56170 18.11904 17.70622 17.32122
#> 17 20.67619 20.09102 19.54529 19.03635 18.56170 18.11904 17.70622 17.32122
#> 18 20.67619 21.31987 21.92018 22.48002 23.00213 23.48906 23.94316 24.36666
#> 19 20.67619 21.31987 21.92018 22.48002 23.00213 23.48906 23.94316 24.36666
#> 20 20.67619 20.09102 19.54529 19.03635 18.56170 18.11904 17.70622 17.32122
#> 21 20.67619 20.70545 20.73274 20.75818 20.78192 20.80405 20.82469 20.84394
#>      2.9347   2.7369   2.5524   2.3804   2.2200   2.0704   1.9308   1.8007
#> 1  20.86189 20.87863 20.89425 20.90881 20.92239 20.93506 20.94687 20.95788
#> 2  20.86189 20.87863 20.89425 20.90881 20.92239 20.93506 20.94687 20.95788
#> 3  24.76162 25.12996 25.47348 25.79384 26.09261 26.37125 26.63110 26.87345
#> 4  20.86189 20.87863 20.89425 20.90881 20.92239 20.93506 20.94687 20.95788
#> 5  16.96216 16.62731 16.31502 16.02378 15.75217 15.49887 15.26263 15.04232
#> 6  20.86189 20.87863 20.89425 20.90881 20.92239 20.93506 20.94687 20.95788
#> 7  24.76162 25.12996 25.47348 25.79384 26.09261 26.37125 26.63110 26.87345
#> 8  24.76162 25.12996 25.47348 25.79384 26.09261 26.37125 26.63110 26.87345
#> 9  20.86189 20.87863 20.89425 20.90881 20.92239 20.93506 20.94687 20.95788
#> 10 16.96216 16.62731 16.31502 16.02378 15.75217 15.49887 15.26263 15.04232
#> 11 16.96216 16.62731 16.31502 16.02378 15.75217 15.49887 15.26263 15.04232
#> 12 16.96216 16.62731 16.31502 16.02378 15.75217 15.49887 15.26263 15.04232
#> 13 24.76162 25.12996 25.47348 25.79384 26.09261 26.37125 26.63110 26.87345
#> 14 24.76162 25.12996 25.47348 25.79384 26.09261 26.37125 26.63110 26.87345
#> 15 16.96216 16.62731 16.31502 16.02378 15.75217 15.49887 15.26263 15.04232
#> 16 16.96216 16.62731 16.31502 16.02378 15.75217 15.49887 15.26263 15.04232
#> 17 16.96216 16.62731 16.31502 16.02378 15.75217 15.49887 15.26263 15.04232
#> 18 24.76162 25.12996 25.47348 25.79384 26.09261 26.37125 26.63110 26.87345
#> 19 24.76162 25.12996 25.47348 25.79384 26.09261 26.37125 26.63110 26.87345
#> 20 16.96216 16.62731 16.31502 16.02378 15.75217 15.49887 15.26263 15.04232
#> 21 20.86189 20.87863 20.89425 20.90881 20.92239 20.93506 20.94687 20.95788
#>      1.6793   1.5662   1.4606   1.3622   1.2704   1.1847   1.1049   1.0304
#> 1  20.96561 20.96561 20.96561 20.96561 20.96561 20.96561 21.39895 21.80862
#> 2  20.96561 20.96561 20.96561 20.96561 20.96561 20.96561 21.19056 21.40321
#> 3  27.04331 27.04331 27.04331 27.04331 27.04331 27.04331 27.02598 27.00918
#> 4  20.96561 20.96561 20.96561 20.96561 20.96561 20.96561 20.91270 20.86267
#> 5  14.88790 14.88790 14.88790 14.88790 14.88790 14.88790 15.34696 15.78134
#> 6  20.96561 20.96561 20.96561 20.96561 20.96561 20.96561 20.71248 20.47316
#> 7  27.04331 27.04331 27.04331 27.04331 27.04331 27.04331 26.31500 25.62602
#> 8  27.04331 27.04331 27.04331 27.04331 27.04331 27.04331 26.34769 25.68962
#> 9  20.96561 20.96561 20.96561 20.96561 20.96561 20.96561 20.72883 20.50495
#> 10 14.88790 14.88790 14.88790 14.88790 14.88790 14.88790 15.06911 15.24080
#> 11 14.88790 14.88790 14.88790 14.88790 14.88790 14.88790 13.72559 12.62710
#> 12 14.88790 14.88790 14.88790 14.88790 14.88790 14.88790 13.79015 12.75270
#> 13 27.04331 27.04331 27.04331 27.04331 27.04331 27.04331 27.12405 27.19996
#> 14 27.04331 27.04331 27.04331 27.04331 27.04331 27.04331 27.42233 27.78025
#> 15 14.88790 14.88790 14.88790 14.88790 14.88790 14.88790 15.35105 15.78929
#> 16 14.88790 14.88790 14.88790 14.88790 14.88790 14.88790 15.02007 15.14541
#> 17 14.88790 14.88790 14.88790 14.88790 14.88790 14.88790 15.01599 15.13746
#> 18 27.04331 27.04331 27.04331 27.04331 27.04331 27.04331 27.17308 27.29535
#> 19 27.04331 27.04331 27.04331 27.04331 27.04331 27.04331 27.68548 28.29218
#> 20 14.88790 14.88790 14.88790 14.88790 14.88790 14.88790 15.56761 16.21060
#> 21 20.96561 20.96561 20.96561 20.96561 20.96561 20.96561 21.27636 21.57014
#>      0.9610   0.8962   0.8358   0.7795   0.7269   0.6779   0.6323   0.5896
#> 1  22.19097 22.47828 22.47030 22.34233 22.22336 22.11228 22.00871 21.91212
#> 2  21.60168 21.75082 21.74285 21.61733 21.50065 21.39180 21.29030 21.19565
#> 3  26.99341 26.98143 26.98608 27.06430 27.13704 27.20509 27.26855 27.32773
#> 4  20.81596 20.78086 20.79028 20.93719 21.07373 21.20106 21.31977 21.43048
#> 5  16.18684 16.49167 16.50003 16.61947 16.73042 16.83343 16.92946 17.01901
#> 6  20.24978 20.08192 20.09136 20.24063 20.37937 20.50882 20.62953 20.74210
#> 7  24.98289 24.49949 24.49841 24.49527 24.49243 24.49032 24.48839 24.48659
#> 8  25.07533 24.61360 24.61252 24.60899 24.60579 24.60334 24.60108 24.59898
#> 9  20.29600 20.13898 20.13105 20.01096 19.89936 19.79543 19.69853 19.60817
#> 10 15.40112 15.52172 15.52431 15.55730 15.58791 15.61615 15.64247 15.66701
#> 11 11.60193 10.83170 10.82862 10.78767 10.74966 10.71450 10.68173 10.65117
#> 12 11.78449 11.05707 11.05399 11.01228 10.97356 10.93771 10.90429 10.87313
#> 13 27.27073 27.32377 27.32841 27.40548 27.47714 27.54414 27.60662 27.66489
#> 14 28.11422 28.36504 28.36966 28.44322 28.51160 28.57542 28.63493 28.69043
#> 15 16.19840 16.50594 16.51430 16.63369 16.74459 16.84756 16.94355 17.03306
#> 16 15.26247 15.35055 15.34736 15.29120 15.23896 15.18999 15.14431 15.10172
#> 17 15.25091 15.33629 15.34467 15.46800 15.58258 15.68913 15.78846 15.88109
#> 18 27.40938 27.49493 27.49379 27.48056 27.46829 27.45703 27.44654 27.43676
#> 19 28.85835 29.28365 29.28246 29.26320 29.24530 29.22858 29.21298 29.19844
#> 20 16.81080 17.26193 17.25870 17.19610 17.13784 17.08303 17.03190 16.98421
#> 21 21.84433 22.05036 22.03082 21.72483 21.44044 21.17520 20.92789 20.69725
#>      0.5499   0.5128   0.4783   0.4460   0.4160   0.3879   0.3618   0.3374
#> 1  21.82204 21.73803 21.65968 21.58661 21.51847 21.45492 21.39565 21.65234
#> 2  21.10737 21.02505 20.94827 20.87667 20.80989 20.74762 20.68954 21.19226
#> 3  27.38292 27.43439 27.48240 27.52717 27.56892 27.60786 27.64417 28.53522
#> 4  21.53373 21.63002 21.71982 21.80356 21.88167 21.95451 22.02244 21.65760
#> 5  17.10253 17.18042 17.25306 17.32080 17.38398 17.44290 17.49785 17.10022
#> 6  20.84709 20.94500 21.03631 21.12146 21.20088 21.27494 21.34402 21.21556
#> 7  24.48491 24.48334 24.48188 24.48052 24.47925 24.47806 24.47696 23.66394
#> 8  24.59701 24.59518 24.59347 24.59188 24.59040 24.58901 24.58772 23.73611
#> 9  19.52390 19.44531 19.37201 19.30366 19.23991 19.18046 19.12501 17.96636
#> 10 15.68990 15.71125 15.73115 15.74972 15.76703 15.78318 15.79824 15.39168
#> 11 10.62267 10.59609 10.57130 10.54818 10.52662 10.50652 10.48777 11.33041
#> 12 10.84407 10.81697 10.79170 10.76813 10.74614 10.72564 10.70652 11.47295
#> 13 27.71923 27.76991 27.81718 27.86126 27.90237 27.94071 27.97646 28.75173
#> 14 28.74219 28.79045 28.83547 28.87745 28.91660 28.95312 28.98717 29.41028
#> 15 17.11654 17.19440 17.26701 17.33473 17.39788 17.45677 17.51170 17.10924
#> 16 15.06200 15.02495 14.99040 14.95818 14.92813 14.90011 14.87397 14.18834
#> 17 15.96748 16.04804 16.12318 16.19325 16.25860 16.31954 16.37638 16.36950
#> 18 27.42764 27.41914 27.41121 27.40381 27.39691 27.39048 27.38448 27.76489
#> 19 29.18488 29.17223 29.16043 29.14943 29.13917 29.12960 29.12068 28.89616
#> 20 16.93974 16.89827 16.85959 16.82352 16.78988 16.75850 16.72924 17.60368
#> 21 20.48215 20.28156 20.09448 19.92001 19.75730 19.60555 19.46403 19.19152
#>      0.3147   0.2935   0.2737   0.2552   0.2380   0.2220   0.2070   0.1931
#> 1  21.65233 21.65232 21.65231 21.65230 21.65230 21.65230 21.65230 21.65230
#> 2  21.19235 21.19241 21.19245 21.19247 21.19249 21.19250 21.19251 21.19251
#> 3  28.53551 28.53570 28.53583 28.53591 28.53596 28.53599 28.53601 28.53602
#> 4  21.65748 21.65740 21.65735 21.65731 21.65729 21.65727 21.65726 21.65725
#> 5  17.09979 17.09950 17.09929 17.09916 17.09907 17.09900 17.09896 17.09894
#> 6  21.21554 21.21553 21.21552 21.21552 21.21551 21.21551 21.21550 21.21550
#> 7  23.66420 23.66439 23.66452 23.66462 23.66468 23.66473 23.66476 23.66478
#> 8  23.73636 23.73653 23.73666 23.73675 23.73681 23.73685 23.73688 23.73690
#> 9  17.96631 17.96629 17.96628 17.96628 17.96628 17.96629 17.96629 17.96630
#> 10 15.39139 15.39118 15.39104 15.39095 15.39089 15.39085 15.39082 15.39080
#> 11 11.33078 11.33103 11.33119 11.33130 11.33137 11.33142 11.33145 11.33147
#> 12 11.47328 11.47351 11.47366 11.47376 11.47382 11.47387 11.47389 11.47391
#> 13 28.75197 28.75213 28.75223 28.75230 28.75234 28.75237 28.75238 28.75239
#> 14 29.41037 29.41043 29.41046 29.41049 29.41050 29.41051 29.41051 29.41051
#> 15 17.10881 17.10851 17.10831 17.10817 17.10808 17.10802 17.10798 17.10795
#> 16 14.18805 14.18786 14.18774 14.18765 14.18760 14.18757 14.18754 14.18753
#> 17 16.36924 16.36906 16.36893 16.36884 16.36878 16.36874 16.36871 16.36869
#> 18 27.76510 27.76524 27.76533 27.76539 27.76543 27.76546 27.76548 27.76549
#> 19 28.89610 28.89607 28.89605 28.89603 28.89602 28.89602 28.89601 28.89601
#> 20 17.60349 17.60336 17.60326 17.60319 17.60314 17.60311 17.60309 17.60307
#> 21 19.19154 19.19157 19.19159 19.19161 19.19163 19.19164 19.19165 19.19166
#>      0.1801   0.1679   0.1566   0.1461   0.1362   0.1270   0.1185   0.1105
#> 1  21.65230 21.65230 21.65230 21.65230 21.65230 21.65230 22.14594 22.14583
#> 2  21.19252 21.19252 21.19252 21.19252 21.19252 21.19252 21.63407 21.63395
#> 3  28.53603 28.53604 28.53604 28.53604 28.53604 28.53604 28.63317 28.63327
#> 4  21.65725 21.65725 21.65725 21.65724 21.65724 21.65724 21.57815 21.57839
#> 5  17.09892 17.09891 17.09890 17.09890 17.09890 17.09890 17.07774 17.07797
#> 6  21.21550 21.21550 21.21550 21.21550 21.21550 21.21550 20.76348 20.76397
#> 7  23.66479 23.66480 23.66481 23.66481 23.66482 23.66482 23.46291 23.46267
#> 8  23.73692 23.73692 23.73693 23.73693 23.73694 23.73694 23.77527 23.77484
#> 9  17.96630 17.96630 17.96630 17.96631 17.96631 17.96631 18.16571 18.16523
#> 10 15.39079 15.39079 15.39078 15.39078 15.39078 15.39078 15.40648 15.40666
#> 11 11.33148 11.33149 11.33149 11.33149 11.33149 11.33149 11.12774 11.12783
#> 12 11.47392 11.47393 11.47393 11.47393 11.47393 11.47394 11.51839 11.51829
#> 13 28.75240 28.75240 28.75240 28.75240 28.75240 28.75240 29.10612 29.10603
#> 14 29.41051 29.41051 29.41051 29.41051 29.41051 29.41051 29.98005 29.97986
#> 15 17.10794 17.10793 17.10792 17.10791 17.10791 17.10791 17.08778 17.08801
#> 16 14.18752 14.18752 14.18751 14.18751 14.18751 14.18751 15.04392 15.04345
#> 17 16.36868 16.36867 16.36867 16.36867 16.36867 16.36866 16.19415 16.19442
#> 18 27.76549 27.76550 27.76550 27.76550 27.76550 27.76550 27.58747 27.58727
#> 19 28.89601 28.89601 28.89601 28.89601 28.89601 28.89601 28.18015 28.18051
#> 20 17.60306 17.60305 17.60305 17.60305 17.60304 17.60304 17.06927 17.06931
#> 21 19.19166 19.19167 19.19167 19.19167 19.19167 19.19167 18.66204 18.66224
#>      0.1030   0.0961   0.0896   0.0836   0.0779   0.0727   0.0678   0.0632
#> 1  22.14580 22.73550 22.73573 22.73596 23.16757 23.16766 23.16774 23.16782
#> 2  21.63392 21.92811 21.92839 21.92867 22.12978 22.12977 22.12976 22.12975
#> 3  28.63338 27.81623 27.81596 27.81568 27.55033 27.55032 27.55031 27.55030
#> 4  21.57854 22.14895 22.14874 22.14856 22.56221 22.56229 22.56237 22.56245
#> 5  17.07808 17.88095 17.88082 17.88070 17.96753 17.96763 17.96772 17.96782
#> 6  20.76432 20.21715 20.21707 20.21699 20.01229 20.01217 20.01205 20.01193
#> 7  23.46251 23.41042 23.41062 23.41079 23.43297 23.43290 23.43284 23.43277
#> 8  23.77454 24.66846 24.66843 24.66839 24.67997 24.68018 24.68038 24.68058
#> 9  18.16486 17.52919 17.52979 17.53036 17.92853 17.92825 17.92798 17.92772
#> 10 15.40677 14.13721 14.13745 14.13767 13.46548 13.46531 13.46515 13.46500
#> 11 11.12794 11.72740 11.72720 11.72702 11.90109 11.90118 11.90126 11.90135
#> 12 11.51826 11.20069 11.20056 11.20042 11.07015 11.07017 11.07018 11.07020
#> 13 29.10600 28.71727 28.71722 28.71717 28.55648 28.55638 28.55628 28.55618
#> 14 29.97974 30.51640 30.51629 30.51619 30.73227 30.73235 30.73243 30.73251
#> 15 17.08811 16.81449 16.81469 16.81490 16.43548 16.43528 16.43509 16.43490
#> 16 15.04310 14.19288 14.19281 14.19271 14.33410 14.33435 14.33459 14.33483
#> 17 16.19457 17.27054 17.27034 17.27017 17.41039 17.41046 17.41053 17.41060
#> 18 27.58712 27.38347 27.38351 27.38355 26.59035 26.59025 26.59016 26.59007
#> 19 28.18074 28.29284 28.29251 28.29220 28.54746 28.54760 28.54774 28.54787
#> 20 17.06928 17.37071 17.37053 17.37039 17.31288 17.31270 17.31253 17.31236
#> 21 18.66241 18.24113 18.24132 18.24150 18.41269 18.41279 18.41289 18.41299
#>      0.0590   0.0550   0.0513   0.0478   0.0446   0.0416   0.0388   0.0362
#> 1  23.16790 23.16798 23.16805 23.16812 23.16818 23.16825 23.16831 23.16837
#> 2  22.12974 22.12972 22.12971 22.12970 22.12969 22.12967 22.12966 22.12965
#> 3  27.55029 27.55028 27.55027 27.55027 27.55026 27.55026 27.55025 27.55025
#> 4  22.56253 22.56260 22.56267 22.56273 22.56280 22.56286 22.56291 22.56297
#> 5  17.96791 17.96800 17.96808 17.96816 17.96824 17.96832 17.96839 17.96846
#> 6  20.01181 20.01170 20.01159 20.01149 20.01139 20.01129 20.01119 20.01110
#> 7  23.43271 23.43265 23.43259 23.43253 23.43248 23.43242 23.43237 23.43232
#> 8  24.68077 24.68096 24.68114 24.68132 24.68150 24.68167 24.68183 24.68199
#> 9  17.92746 17.92721 17.92696 17.92672 17.92649 17.92626 17.92604 17.92582
#> 10 13.46485 13.46470 13.46456 13.46443 13.46430 13.46418 13.46406 13.46394
#> 11 11.90143 11.90151 11.90158 11.90165 11.90173 11.90179 11.90186 11.90192
#> 12 11.07022 11.07023 11.07025 11.07027 11.07029 11.07030 11.07032 11.07034
#> 13 28.55609 28.55600 28.55591 28.55582 28.55574 28.55566 28.55558 28.55550
#> 14 30.73258 30.73265 30.73272 30.73279 30.73285 30.73292 30.73298 30.73303
#> 15 16.43472 16.43454 16.43437 16.43420 16.43404 16.43388 16.43372 16.43357
#> 16 14.33507 14.33530 14.33552 14.33574 14.33596 14.33617 14.33638 14.33659
#> 17 17.41067 17.41073 17.41079 17.41085 17.41090 17.41095 17.41100 17.41105
#> 18 26.58998 26.58990 26.58982 26.58975 26.58968 26.58961 26.58954 26.58948
#> 19 28.54800 28.54812 28.54824 28.54836 28.54847 28.54858 28.54869 28.54880
#> 20 17.31220 17.31204 17.31188 17.31172 17.31157 17.31142 17.31127 17.31113
#> 21 18.41309 18.41318 18.41328 18.41337 18.41345 18.41354 18.41363 18.41371
#>      0.0337   0.0315   0.0293   0.0274   0.0255   0.0238   0.0222   0.0207
#> 1  23.16843 23.16848 23.16853 23.16858 23.16863 23.16868 23.16872 23.16876
#> 2  22.12964 22.12962 22.12961 22.12960 22.12958 22.12957 22.12956 22.12954
#> 3  27.55025 27.55025 27.55025 27.55026 27.55026 27.55026 27.55027 27.55027
#> 4  22.56302 22.56308 22.56312 22.56317 22.56322 22.56326 22.56330 22.56334
#> 5  17.96853 17.96859 17.96866 17.96872 17.96877 17.96883 17.96888 17.96893
#> 6  20.01100 20.01092 20.01083 20.01075 20.01067 20.01059 20.01051 20.01044
#> 7  23.43227 23.43223 23.43218 23.43214 23.43209 23.43205 23.43201 23.43197
#> 8  24.68215 24.68230 24.68245 24.68259 24.68273 24.68287 24.68300 24.68313
#> 9  17.92561 17.92541 17.92521 17.92501 17.92482 17.92464 17.92446 17.92429
#> 10 13.46383 13.46373 13.46362 13.46353 13.46343 13.46334 13.46326 13.46317
#> 11 11.90198 11.90204 11.90210 11.90216 11.90221 11.90226 11.90231 11.90236
#> 12 11.07036 11.07037 11.07039 11.07041 11.07043 11.07044 11.07046 11.07048
#> 13 28.55543 28.55536 28.55529 28.55522 28.55515 28.55509 28.55503 28.55497
#> 14 30.73309 30.73314 30.73320 30.73325 30.73329 30.73334 30.73339 30.73343
#> 15 16.43342 16.43328 16.43314 16.43301 16.43288 16.43275 16.43263 16.43251
#> 16 14.33679 14.33698 14.33717 14.33736 14.33754 14.33772 14.33789 14.33806
#> 17 17.41109 17.41113 17.41117 17.41121 17.41125 17.41128 17.41131 17.41134
#> 18 26.58942 26.58936 26.58931 26.58926 26.58921 26.58916 26.58911 26.58907
#> 19 28.54890 28.54900 28.54909 28.54918 28.54927 28.54936 28.54944 28.54953
#> 20 17.31099 17.31085 17.31072 17.31059 17.31046 17.31034 17.31021 17.31009
#> 21 18.41379 18.41387 18.41395 18.41402 18.41410 18.41417 18.41424 18.41431
#>      0.0193   0.0180   0.0168   0.0157   0.0146   0.0136   0.0127   0.0118
#> 1  23.16880 23.16884 23.16888 23.16892 23.16895 23.16898 23.16901 23.16904
#> 2  22.12953 22.12952 22.12950 22.12949 22.12948 22.12947 22.12945 22.12944
#> 3  27.55028 27.55028 27.55029 27.55029 27.55030 27.55031 27.55031 27.55032
#> 4  22.56338 22.56341 22.56345 22.56348 22.56351 22.56354 22.56357 22.56359
#> 5  17.96898 17.96903 17.96907 17.96912 17.96916 17.96920 17.96924 17.96927
#> 6  20.01037 20.01030 20.01023 20.01017 20.01011 20.01005 20.00999 20.00993
#> 7  23.43194 23.43190 23.43187 23.43183 23.43180 23.43177 23.43174 23.43171
#> 8  24.68326 24.68338 24.68350 24.68361 24.68372 24.68383 24.68393 24.68404
#> 9  17.92412 17.92395 17.92379 17.92364 17.92349 17.92334 17.92320 17.92306
#> 10 13.46309 13.46302 13.46294 13.46287 13.46280 13.46274 13.46268 13.46262
#> 11 11.90240 11.90245 11.90249 11.90253 11.90257 11.90261 11.90265 11.90268
#> 12 11.07050 11.07051 11.07053 11.07055 11.07057 11.07058 11.07060 11.07061
#> 13 28.55491 28.55486 28.55481 28.55475 28.55471 28.55466 28.55461 28.55457
#> 14 30.73347 30.73351 30.73355 30.73359 30.73362 30.73366 30.73369 30.73372
#> 15 16.43239 16.43228 16.43217 16.43206 16.43196 16.43186 16.43176 16.43166
#> 16 14.33823 14.33839 14.33855 14.33870 14.33885 14.33900 14.33914 14.33928
#> 17 17.41137 17.41140 17.41143 17.41145 17.41147 17.41149 17.41152 17.41153
#> 18 26.58903 26.58899 26.58895 26.58892 26.58888 26.58885 26.58882 26.58879
#> 19 28.54961 28.54968 28.54976 28.54983 28.54990 28.54996 28.55003 28.55009
#> 20 17.30998 17.30986 17.30975 17.30964 17.30954 17.30943 17.30933 17.30924
#> 21 18.41437 18.41444 18.41450 18.41456 18.41462 18.41468 18.41474 18.41479
#>      0.0110   0.0103   0.0096   0.0090   0.0084   0.0078   0.0073   0.0068
#> 1  23.16907 23.16910 23.16912 23.16915 23.16917 23.16919 23.16921 23.16923
#> 2  22.12943 22.12942 22.12941 22.12939 22.12938 22.12937 22.12936 22.12935
#> 3  27.55033 27.55034 27.55035 27.55036 27.55037 27.55037 27.55038 27.55039
#> 4  22.56362 22.56364 22.56366 22.56369 22.56371 22.56373 22.56374 22.56376
#> 5  17.96931 17.96934 17.96938 17.96941 17.96944 17.96946 17.96949 17.96952
#> 6  20.00988 20.00982 20.00977 20.00972 20.00968 20.00963 20.00959 20.00954
#> 7  23.43168 23.43165 23.43163 23.43160 23.43158 23.43155 23.43153 23.43151
#> 8  24.68413 24.68423 24.68432 24.68441 24.68450 24.68458 24.68466 24.68474
#> 9  17.92293 17.92280 17.92267 17.92255 17.92243 17.92231 17.92220 17.92209
#> 10 13.46256 13.46251 13.46246 13.46241 13.46236 13.46232 13.46227 13.46223
#> 11 11.90272 11.90275 11.90278 11.90281 11.90284 11.90287 11.90290 11.90292
#> 12 11.07063 11.07065 11.07066 11.07068 11.07069 11.07071 11.07072 11.07074
#> 13 28.55452 28.55448 28.55444 28.55440 28.55437 28.55433 28.55429 28.55426
#> 14 30.73375 30.73378 30.73381 30.73383 30.73386 30.73388 30.73391 30.73393
#> 15 16.43157 16.43149 16.43140 16.43132 16.43124 16.43116 16.43108 16.43101
#> 16 14.33941 14.33954 14.33967 14.33980 14.33992 14.34004 14.34015 14.34026
#> 17 17.41155 17.41157 17.41158 17.41160 17.41161 17.41162 17.41164 17.41165
#> 18 26.58876 26.58874 26.58871 26.58869 26.58867 26.58865 26.58863 26.58861
#> 19 28.55015 28.55021 28.55027 28.55032 28.55038 28.55043 28.55048 28.55052
#> 20 17.30914 17.30905 17.30896 17.30887 17.30878 17.30870 17.30862 17.30854
#> 21 18.41485 18.41490 18.41495 18.41500 18.41505 18.41509 18.41514 18.41518
#>      0.0063   0.0059   0.0055   0.0051
#> 1  23.16925 23.16927 23.16929 23.16930
#> 2  22.12934 22.12933 22.12932 22.12930
#> 3  27.55040 27.55041 27.55042 27.55043
#> 4  22.56378 22.56379 22.56381 22.56382
#> 5  17.96954 17.96956 17.96959 17.96961
#> 6  20.00950 20.00946 20.00942 20.00939
#> 7  23.43149 23.43147 23.43145 23.43143
#> 8  24.68482 24.68489 24.68496 24.68503
#> 9  17.92199 17.92189 17.92179 17.92170
#> 10 13.46219 13.46216 13.46212 13.46209
#> 11 11.90295 11.90297 11.90299 11.90301
#> 12 11.07075 11.07077 11.07078 11.07079
#> 13 28.55423 28.55420 28.55417 28.55414
#> 14 30.73395 30.73397 30.73399 30.73401
#> 15 16.43094 16.43087 16.43081 16.43075
#> 16 14.34037 14.34048 14.34058 14.34068
#> 17 17.41166 17.41166 17.41167 17.41168
#> 18 26.58859 26.58858 26.58856 26.58855
#> 19 28.55057 28.55061 28.55066 28.55070
#> 20 17.30846 17.30839 17.30832 17.30824
#> 21 18.41522 18.41526 18.41531 18.41534
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 22.8 21.4 18.7 18.1 24.4 22.8 17.8 15.2 10.4 14.7 32.4 33.9 15.2
#> [16] 13.3 19.2 26.0 30.4 15.8 19.7
#> 
#> $X
#>               am       carb         cyl        disp        drat       gear
#>  [1,]  1.1547005  0.9581403 -0.05643326 -0.57174039  0.62250989  0.3175003
#>  [2,]  1.1547005  0.9581403 -0.05643326 -0.57174039  0.62250989  0.3175003
#>  [3,]  1.1547005 -1.1976753 -1.24153183 -1.00022197  0.51588095  0.3175003
#>  [4,] -0.8660254 -1.1976753 -0.05643326  0.23578258 -1.12620468 -1.0160010
#>  [5,] -0.8660254 -0.4790701  1.12866530  1.07626568 -0.97692417 -1.0160010
#>  [6,] -0.8660254 -1.1976753 -0.05643326 -0.03613842 -1.80862988 -1.0160010
#>  [7,] -0.8660254 -0.4790701 -1.24153183 -0.68133279  0.17466835  0.3175003
#>  [8,] -0.8660254 -0.4790701 -1.24153183 -0.72994897  0.66516146  0.3175003
#>  [9,] -0.8660254  0.9581403 -0.05643326 -0.50911616  0.66516146  0.3175003
#> [10,] -0.8660254  0.2395351  1.12866530  0.38245512 -1.14753047 -1.0160010
#> [11,] -0.8660254  0.9581403  1.12866530  1.90026871 -1.29681098 -1.0160010
#> [12,] -0.8660254  0.9581403  1.12866530  1.73546810 -0.80631787 -1.0160010
#> [13,]  1.1547005 -1.1976753 -1.24153183 -1.24165486  1.00637406  0.3175003
#> [14,]  1.1547005 -1.1976753 -1.24153183 -1.30427909  1.30493509  0.3175003
#> [15,] -0.8660254 -0.4790701  1.12866530  0.61482398 -0.97692417 -1.0160010
#> [16,] -0.8660254  0.9581403  1.12866530  0.99386537  0.25997150 -1.0160010
#> [17,] -0.8660254 -0.4790701  1.12866530  1.40586689 -1.12620468 -1.0160010
#> [18,]  1.1547005 -0.4790701 -1.24153183 -0.89886959  1.75277662  1.6510017
#> [19,]  1.1547005 -0.4790701 -1.24153183 -1.10651836  0.34527465  1.6510017
#> [20,]  1.1547005  0.9581403  1.12866530  1.00210540  1.30493509  1.6510017
#> [21,]  1.1547005  2.3953507 -0.05643326 -0.69534085  0.02538784  1.6510017
#>               hp        qsec         vs            wt
#>  [1,] -0.5141103 -0.77283821 -0.8660254 -0.5972263465
#>  [2,] -0.5141103 -0.47851900 -0.8660254 -0.3304250546
#>  [3,] -0.7999940  0.35713735  1.1547005 -0.9111102192
#>  [4,] -0.5141103  0.79336047  1.1547005  0.0253100012
#>  [5,]  0.5789746 -0.47851900 -0.8660254  0.2607229057
#>  [6,] -0.5981937  1.20330509  1.1547005  0.2816484973
#>  [7,] -1.3213114  1.08767968  1.1547005 -0.0008469882
#>  [8,] -0.7663606  2.61183276  1.1547005 -0.0426981713
#>  [9,] -0.2954933  0.50955265  1.1547005  0.2607229057
#> [10,]  0.6630581  0.03653963 -0.8660254  0.6164579615
#> [11,]  1.2516422 -0.05806297 -0.8660254  2.3365415842
#> [12,]  1.5038926 -0.26829099 -0.8660254  2.2538854977
#> [13,] -1.2540446  0.80912757  1.1547005 -1.0366637683
#> [14,] -1.2708613  1.03512268  1.1547005 -1.4185558135
#> [15,]  0.1585574 -0.33135939 -0.8660254  0.2554915079
#> [16,]  1.7561430 -1.32468674 -0.8660254  0.6792347361
#> [17,]  0.5789746 -0.46275190 -0.8660254  0.6844661340
#> [18,] -0.8336274 -0.64670140 -0.8660254 -1.0994405429
#> [19,] -0.4636602 -0.54158740  1.1547005 -1.7554578369
#> [20,]  2.0756601 -1.80295546 -0.8660254 -0.0217725797
#> [21,]  0.5789746 -1.27738544 -0.8660254 -0.4402844101
#> attr(,"center")
#>  [1]   0.4285714   2.6666667   6.0952381 229.3857143   3.6080952   3.7619048
#>  [7] 140.5714286  17.9304762   0.4285714   3.1908095
#> attr(,"scale")
#>  [1]   0.4948717   1.3915847   1.6876233 121.3587766   0.4689159   0.7499055
#>  [7]  59.4647326   1.9026960   0.4948717   0.9557675
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 30
#> 
#> $lambda.min
#> [1] 0.6779496
#> 
#> $null.dev
#> [1] 36.20277
#> 
#> $Bias
#> [1] 8.463182
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 5.375215