Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 34.368173 33.620472 31.457070 28.226662 25.455103 23.157052 21.263250
#>   [8] 19.713973 18.457751 17.450261 16.653372 16.034301 15.564900 15.221021
#>  [15] 14.981975 14.798207 14.467671 14.253794 14.139038 14.082033 14.006120
#>  [22] 13.715750 13.304925 12.587024 10.882711 10.080350  9.058969  8.987781
#>  [29]  8.927513  8.857177  9.183080  9.429031  9.474598 10.195979 10.197385
#>  [36] 10.844638 11.502457 12.073801 11.883725 11.560701 11.412249 15.320008
#>  [43] 15.315865 14.154112 14.587800 19.822031 20.141696 20.580381 21.286316
#>  [50] 24.925649 34.838771 34.841404 34.901376 35.116918 35.128187 35.141749
#>  [57] 35.002106 35.040983 35.042553 35.087071 35.088498 35.185213 36.260590
#>  [64] 36.123665 36.126398 36.280672 36.450072 36.564192 36.753977 36.931997
#>  [71] 37.258091 37.259997 37.261759 37.414154 37.436405 37.437807 37.708397
#>  [78] 37.709710 37.710941 37.712098 37.748157 37.748999 37.725770 37.726556
#>  [85] 37.727305 37.728022 37.860188 37.861346 37.862475 37.863575 37.861992
#>  [92] 37.863030 37.864099 37.904824 37.994695 37.986682 37.986694 37.986700
#>  [99] 37.986750 37.984512
#> 
#> $cvse
#>   [1] 12.777084 12.788181 12.322164 11.146180 10.040414  9.053180  8.170192
#>   [8]  7.380241  6.674750  6.047385  5.493708  5.010829  4.597015  4.251211
#>  [15]  3.972447  3.751574  3.531114  3.352839  3.256538  3.264763  3.273287
#>  [22]  3.230399  3.119144  2.990004  2.907047  2.929736  2.728789  2.727985
#>  [29]  2.728099  2.705093  2.848432  2.978891  2.966377  2.881453  2.865010
#>  [36]  3.267804  3.001693  3.005865  3.003114  2.994328  2.973016  4.966479
#>  [43]  4.963909  5.022539  5.029599  9.339150  9.396358  9.420124  9.364223
#>  [50] 10.026863 14.249388 14.250689 14.245525 14.231731 14.231917 14.231839
#>  [57] 14.251013 14.248941 14.249859 14.246085 14.246917 14.236995 14.636636
#>  [64] 14.310034 14.310233 14.314840 14.296679 14.080892 14.078124 14.078995
#>  [71] 14.078250 14.078958 14.079595 14.097400 14.095359 14.095836 14.071134
#>  [78] 14.071523 14.071881 14.072213 14.076623 14.076925 14.080078 14.080345
#>  [85] 14.080598 14.080841 14.152280 14.152764 14.153238 14.153703 14.154479
#>  [92] 14.154926 14.155358 14.150930 14.144758 14.134643 14.134244 14.133849
#>  [99] 14.133453 14.133313
#> 
#> $fold
#>  [1]  6  1  2  5 10  8  9  2  9  7  5  4  1  1  8  3  7  3 10  4  6
#> 
#> $lambda
#>   [1] 4.874382678 4.545865600 4.239489473 3.953762071 3.687291741 3.438780618
#>   [7] 3.207018314 2.990876013 2.789300980 2.601311429 2.425991745 2.262488021
#>  [13] 2.110003901 1.967796700 1.835173788 1.711489217 1.596140572 1.488566039
#>  [19] 1.388241670 1.294678828 1.207421808 1.126045619 1.050153913 0.979377054
#>  [25] 0.913370319 0.851812216 0.794402924 0.740862826 0.690931151 0.644364704
#>  [31] 0.600936679 0.560435558 0.522664077 0.487438268 0.454586560 0.423948947
#>  [37] 0.395376207 0.368729174 0.343878062 0.320701831 0.299087601 0.278930098
#>  [43] 0.260131143 0.242599175 0.226248802 0.211000390 0.196779670 0.183517379
#>  [49] 0.171148922 0.159614057 0.148856604 0.138824167 0.129467883 0.120742181
#>  [55] 0.112604562 0.105015391 0.097937705 0.091337032 0.085181222 0.079440292
#>  [61] 0.074086283 0.069093115 0.064436470 0.060093668 0.056043556 0.052266408
#>  [67] 0.048743827 0.045458656 0.042394895 0.039537621 0.036872917 0.034387806
#>  [73] 0.032070183 0.029908760 0.027893010 0.026013114 0.024259917 0.022624880
#>  [79] 0.021100039 0.019677967 0.018351738 0.017114892 0.015961406 0.014885660
#>  [85] 0.013882417 0.012946788 0.012074218 0.011260456 0.010501539 0.009793771
#>  [91] 0.009133703 0.008518122 0.007944029 0.007408628 0.006909312 0.006443647
#>  [97] 0.006009367 0.005604356 0.005226641 0.004874383
#> 
#> $fit
#> $beta
#>               4.8744     4.5459     4.2395    3.9538    3.6873    3.4388
#> (Intercept) 19.48095 21.1068666 22.6231996 24.037337 25.356166 26.586110
#> am           0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> wt           0.00000 -0.4816708 -0.9308786 -1.349811 -1.740509 -2.104876
#>                3.2070    2.9909    2.7893    2.6013    2.4260    2.2625
#> (Intercept) 27.733161 28.802904 29.800550 30.730957 31.598659 32.407880
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.444685 -2.761592 -3.057141 -3.332771 -3.589824 -3.829552
#>                2.1100    1.9678    1.8352    1.7115    1.5961    1.4886
#> (Intercept) 33.162563 33.866382 34.522767 35.134913 35.563988 35.563988
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -4.053124 -4.261628 -4.456079 -4.637425 -4.764537 -4.764537
#>                  1.3882    1.2947     1.2074     1.1260    1.0502     0.9794
#> (Intercept) 34.34728494 32.820824 31.3972405 30.0696020 28.831442 27.6767297
#> am           0.00000000  0.000000  0.0000000  0.0000000  0.000000  0.0000000
#> carb         0.00000000  0.000000  0.0000000  0.0000000  0.000000  0.0000000
#> cyl          0.00000000  0.000000  0.0000000  0.0000000  0.000000  0.0000000
#> disp         0.00000000  0.000000  0.0000000  0.0000000  0.000000  0.0000000
#> drat         0.00000000  0.000000  0.0000000  0.0000000  0.000000  0.0000000
#> gear         0.00000000  0.000000  0.0000000  0.0000000  0.000000  0.0000000
#> hp           0.00000000  0.000000  0.0000000  0.0000000  0.000000  0.0000000
#> qsec         0.06839299  0.154198  0.2342201  0.3088489  0.378448  0.4433564
#> vs           0.00000000  0.000000  0.0000000  0.0000000  0.000000  0.0000000
#> wt          -4.76164517 -4.758017 -4.7546329 -4.7514771 -4.748534 -4.7457893
#>                 0.9134     0.8518     0.7944     0.7409     0.6909    0.6444
#> (Intercept) 26.5998412 25.5955314 24.6589087 23.7854112 22.9707845 22.211061
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> qsec         0.5038901  0.5603441  0.6129933  0.6620941  0.7078857  0.750591
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> wt          -4.7432295 -4.7408423 -4.7386159 -4.7365396 -4.7346032 -4.732797
#>                 0.6009     0.5604     0.5227     0.4874    0.4546       0.4239
#> (Intercept) 21.5025402 20.8417714 20.2255362 19.6591285  3.089302  3.154076067
#> am           0.0000000  0.0000000  0.0000000  0.0000000  3.679895  3.695128410
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.000000 -0.009248711
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000000
#> qsec         0.7904182  0.8275611  0.8622008  0.8940395  1.539796  1.535692881
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000000
#> wt          -4.7311132 -4.7295426 -4.7280778 -4.7267314 -3.609230 -3.600076666
#>                 0.3954     0.3687     0.3439     0.3207     0.2991     0.2789
#> (Intercept)  6.3162866  7.9835896  7.9803206  7.9778077  7.9758796  7.9744025
#> am           4.3618072  4.7112683  4.7121389  4.7128047  4.7133135  4.7137019
#> carb        -0.4367296 -0.6614815 -0.6615285 -0.6615636 -0.6615898 -0.6616095
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         1.3394923  1.2361531  1.2362729  1.2363652  1.2364362  1.2364906
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -3.1883811 -2.9722478 -2.9719603 -2.9717409 -2.9715736 -2.9714459
#>                 0.2601       0.2426     0.2262     0.2110    0.1968     0.1835
#> (Intercept)  7.9732722  7.940885201  6.9364730  6.0074380  5.270365  5.2736718
#> am           4.7139985  4.711752272  4.4814173  4.2650620  4.093437  4.0927115
#> carb        -0.6616243 -0.662078677 -0.6950799 -0.7257964 -0.750367 -0.7503834
#> cyl          0.0000000  0.000000000  0.0000000  0.0000000  0.000000  0.0000000
#> disp         0.0000000  0.000000000  0.0000000  0.0000000  0.000000  0.0000000
#> drat         0.0000000  0.008700852  0.4547101  0.8703839  1.200643  1.2006807
#> gear         0.0000000  0.000000000  0.0000000  0.0000000  0.000000  0.0000000
#> hp           0.0000000  0.000000000  0.0000000  0.0000000  0.000000  0.0000000
#> qsec         1.2365323  1.236316650  1.1981021  1.1622176  1.133639  1.1335010
#> vs           0.0000000  0.000000000  0.0000000  0.0000000  0.000000  0.0000000
#> wt          -2.9713486 -2.969009999 -2.8793546 -2.7963291 -2.730200 -2.7304000
#>                 0.1711     0.1596    0.1489    0.1388     0.1295     0.1207
#> (Intercept)  5.2761279  5.2779475  5.279294  5.280289 21.0581631 21.0638947
#> am           4.0921738  4.0917754  4.091480  4.091261  3.9206149  3.9200890
#> carb        -0.7503895 -0.7503897 -0.750387 -0.750383 -0.7540228 -0.7540633
#> cyl          0.0000000  0.0000000  0.000000  0.000000 -0.8115631 -0.8118398
#> disp         0.0000000  0.0000000  0.000000  0.000000  0.0000000  0.0000000
#> drat         1.2006923  1.2006903  1.200682  1.200671  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.000000  0.000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.000000  0.000000  0.0000000  0.0000000
#> qsec         1.1334021  1.1333313  1.133281  1.133244  0.6748251  0.6745827
#> vs           0.0000000  0.0000000  0.000000  0.000000  0.0000000  0.0000000
#> wt          -2.7305566 -2.7306781 -2.730772 -2.730844 -2.2014590 -2.2012697
#>                 0.1126      0.1050     0.0979     0.0913     0.0852     0.0794
#> (Intercept) 21.0688400 21.04962856 17.4672199 17.4612736 17.4560760 17.4515367
#> am           3.9196341  3.91569535  3.7194217  3.7198257  3.7201826  3.7204970
#> carb        -0.7540982 -0.75460358 -0.7857040 -0.7856679 -0.7856369 -0.7856103
#> cyl         -0.8120782 -0.81222516 -0.6762036 -0.6759023 -0.6756401 -0.6754122
#> disp         0.0000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.00809676  0.6359109  0.6360135  0.6361040  0.6361838
#> gear         0.0000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.6743736  0.67372773  0.7131029  0.7133375  0.7135423  0.7137209
#> vs           0.0000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.2011072 -2.19927684 -2.2035723 -2.2037929 -2.2039825 -2.2041457
#>                 0.0741     0.0691     0.0644     0.0601     0.0560     0.0523
#> (Intercept) 19.5490063 19.5555596 19.5614321 19.5666920 19.5714013 19.5756158
#> am           4.0152065  4.0150127  4.0148349  4.0146724  4.0145240  4.0143888
#> carb        -0.8608147 -0.8608953 -0.8609669 -0.8610305 -0.8610870 -0.8611371
#> cyl         -0.6614427 -0.6617304 -0.6619886 -0.6622202 -0.6624278 -0.6626138
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.5757940  0.5756567  0.5755339  0.5754241  0.5753260  0.5752383
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.5472927  0.5470031  0.5467440  0.5465123  0.5463051  0.5461200
#> vs           0.7573068  0.7575619  0.7577874  0.7579866  0.7581628  0.7583186
#> wt          -1.9676769 -1.9673491 -1.9670570 -1.9667968 -1.9665650 -1.9663585
#>                 0.0487     0.0455     0.0424       0.0395       0.0369
#> (Intercept) 19.5793860 19.5827577 19.5857719 23.564418601 23.566640781
#> am           4.0142659  4.0141543  4.0140530  4.566088962  4.566388917
#> carb        -0.8611816 -0.8612212 -0.8612563 -1.134361476 -1.134510177
#> cyl         -0.6627803 -0.6629294 -0.6630628 -0.542049742 -0.542003857
#> disp         0.0000000  0.0000000  0.0000000 -0.007690077 -0.007693512
#> drat         0.5751600  0.5750901  0.5750277  0.431083139  0.431038827
#> gear         0.0000000  0.0000000  0.0000000  0.000000000  0.000000000
#> hp           0.0000000  0.0000000  0.0000000  0.000000000  0.000000000
#> qsec         0.5459546  0.5458069  0.5456750  0.242205079  0.242031998
#> vs           0.7584563  0.7585782  0.7586860  1.258733280  1.259062824
#> wt          -1.9661746 -1.9660109 -1.9658651 -0.944813724 -0.944291785
#>                   0.0344       0.0321       0.0299       0.0279       0.0260
#> (Intercept) 23.568806238 23.570916421 23.572972740 23.574976570 23.576929251
#> am           4.566681214  4.566966051  4.567243617  4.567514098  4.567777675
#> carb        -1.134655081 -1.134796287 -1.134933889 -1.135067978 -1.135198645
#> cyl         -0.541959144 -0.541915573 -0.541873113 -0.541831737 -0.541791418
#> disp        -0.007696859 -0.007700121 -0.007703299 -0.007706397 -0.007709415
#> drat         0.430995647  0.430953568  0.430912564  0.430872606  0.430833669
#> gear         0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> hp           0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> qsec         0.241863335  0.241698977  0.241538814  0.241382739  0.241230649
#> vs           1.259383955  1.259696890  1.260001837  1.260299000  1.260588577
#> wt          -0.943783169 -0.943287536 -0.942804554 -0.942333901 -0.941875261
#>                   0.0243        0.0226       0.0211       0.0197       0.0184
#> (Intercept) 23.578832089 23.5769323787 18.842860536 18.838396477 18.834043775
#> am           4.568034524  4.5678181394  4.403066454  4.402729791  4.402400280
#> carb        -1.135325976 -1.1356080159 -1.166011361 -1.165885986 -1.165763097
#> cyl         -0.541752127 -0.5416283105 -0.431191915 -0.431171922 -0.431152908
#> disp        -0.007712357 -0.0077180269 -0.007631168 -0.007627894 -0.007624687
#> drat         0.430795725  0.4306671855  0.308104124  0.308091068  0.308078656
#> gear         0.000000000  0.0009281101  0.438512323  0.438711389  0.438904915
#> hp           0.000000000  0.0000000000  0.000000000  0.000000000  0.000000000
#> qsec         0.241082441  0.2409209846  0.428857124  0.429114142  0.429365091
#> vs           1.260870763  1.2610173675  0.924243103  0.923783577  0.923334583
#> wt          -0.941428328 -0.9404751876 -1.010724469 -1.011255589 -1.011775364
#>                   0.0171       0.0160       0.0149       0.0139       0.0129
#> (Intercept) 18.829799243 18.825659808 18.821622505 18.817684472 18.813842946
#> am           4.402077782  4.401762161  4.401453284  4.401151018  4.400855232
#> carb        -1.165642656 -1.165524627 -1.165408973 -1.165295657 -1.165184640
#> cyl         -0.431134818 -0.431117601 -0.431101211 -0.431085602 -0.431070732
#> disp        -0.007621547 -0.007618474 -0.007615464 -0.007612518 -0.007609634
#> drat         0.308066853  0.308055624  0.308044938  0.308034766  0.308025080
#> gear         0.439093090  0.439276097  0.439454107  0.439627288  0.439795798
#> hp           0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> qsec         0.429610126  0.429849398  0.430083053  0.430311233  0.430534076
#> vs           0.922895869  0.922467190  0.922048308  0.921638990  0.921239011
#> wt          -1.012284017 -1.012781767 -1.013268831 -1.013745421 -1.014211747
#>                   0.0121       0.0113       0.0105       0.0098       0.0091
#> (Intercept) 18.810095258 18.806438829 18.436604479 18.438198169 18.439814990
#> am           4.400565798  4.400282589  4.378728514  4.379037377  4.379339896
#> carb        -1.165075885 -1.164969355 -1.188983739 -1.189151169 -1.189313723
#> cyl         -0.431056561 -0.431043051 -0.449213242 -0.449155925 -0.449101597
#> disp        -0.007606812 -0.007604049 -0.008131225 -0.008135082 -0.008138834
#> drat         0.308015853  0.308007061  0.309339777  0.309298806  0.309259800
#> gear         0.439959790  0.440119409  0.443695485  0.443712623  0.443724708
#> hp           0.000000000  0.000000000  0.001612322  0.001613198  0.001614036
#> qsec         0.430751714  0.430964280  0.454891207  0.454745816  0.454601519
#> vs           0.920848149  0.920466191  0.844626635  0.844903419  0.845177842
#> wt          -1.014668016 -1.015114432 -1.003356556 -1.002824182 -1.002304601
#>                   0.0085       0.0079       0.0074       0.0069       0.0064
#> (Intercept) 18.441451720 18.443105307 18.444772870 18.446451690 18.448139207
#> am           4.379636127  4.379926137  4.380210002  4.380487804  4.380759633
#> carb        -1.189471566 -1.189624854 -1.189773740 -1.189918371 -1.190058891
#> cyl         -0.449050135 -0.449001413 -0.448955311 -0.448911712 -0.448870500
#> disp        -0.008142483 -0.008146034 -0.008149488 -0.008152851 -0.008156123
#> drat         0.309222689  0.309187401  0.309153864  0.309122007  0.309091760
#> gear         0.443732054  0.443734968  0.443733738  0.443728644  0.443719951
#> hp           0.001614836  0.001615603  0.001616337  0.001617041  0.001617717
#> qsec         0.454458400  0.454316532  0.454175985  0.454036821  0.453899095
#> vs           0.845449715  0.845718868  0.845985153  0.846248443  0.846508624
#> wt          -1.001797490 -1.001302531 -1.000819416 -1.000347839 -0.999887505
#>                   0.0060       0.0056       0.0052       0.0049
#> (Intercept) 18.449833016 18.451530858 18.453230616 18.454930312
#> am           4.381025582  4.381285750  4.381540237  4.381789146
#> carb        -1.190195435 -1.190328137 -1.190457124 -1.190582520
#> cyl         -0.448831563 -0.448794794 -0.448760087 -0.448727342
#> disp        -0.008159309 -0.008162411 -0.008165431 -0.008168372
#> drat         0.309063055  0.309035825  0.309010004  0.308985528
#> gear         0.443707914  0.443692772  0.443674757  0.443654086
#> hp           0.001618365  0.001618988  0.001619587  0.001620162
#> qsec         0.453762858  0.453628156  0.453495029  0.453363512
#> vs           0.846765600  0.847019287  0.847269616  0.847516529
#> wt          -0.999438124 -0.998999414 -0.998571100 -0.998152914
#> 
#> $iter
#>   [1]   0   5   3   3   3   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2  42  18
#>  [37]  78  26   1   1   1   1   1   6  33  33  22   1   1   1   1   1  52   1
#>  [55]   1   7  44   1   1   1  52   1   1   1   1   1   1   1   1 158   1   1
#>  [73]   1   1   1   1   1   2 109   1   1   1   1   1   1   1   1   1  72   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.874382678 4.545865600 4.239489473 3.953762071 3.687291741 3.438780618
#>   [7] 3.207018314 2.990876013 2.789300980 2.601311429 2.425991745 2.262488021
#>  [13] 2.110003901 1.967796700 1.835173788 1.711489217 1.596140572 1.488566039
#>  [19] 1.388241670 1.294678828 1.207421808 1.126045619 1.050153913 0.979377054
#>  [25] 0.913370319 0.851812216 0.794402924 0.740862826 0.690931151 0.644364704
#>  [31] 0.600936679 0.560435558 0.522664077 0.487438268 0.454586560 0.423948947
#>  [37] 0.395376207 0.368729174 0.343878062 0.320701831 0.299087601 0.278930098
#>  [43] 0.260131143 0.242599175 0.226248802 0.211000390 0.196779670 0.183517379
#>  [49] 0.171148922 0.159614057 0.148856604 0.138824167 0.129467883 0.120742181
#>  [55] 0.112604562 0.105015391 0.097937705 0.091337032 0.085181222 0.079440292
#>  [61] 0.074086283 0.069093115 0.064436470 0.060093668 0.056043556 0.052266408
#>  [67] 0.048743827 0.045458656 0.042394895 0.039537621 0.036872917 0.034387806
#>  [73] 0.032070183 0.029908760 0.027893010 0.026013114 0.024259917 0.022624880
#>  [79] 0.021100039 0.019677967 0.018351738 0.017114892 0.015961406 0.014885660
#>  [85] 0.013882417 0.012946788 0.012074218 0.011260456 0.010501539 0.009793771
#>  [91] 0.009133703 0.008518122 0.007944029 0.007408628 0.006909312 0.006443647
#>  [97] 0.006009367 0.005604356 0.005226641 0.004874383
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 34
#> 
#> $loss
#>   [1] 662.65238 566.86873 486.73147 419.98895 364.69723 319.17897 281.98786
#>   [8] 251.87762 227.77523 208.75758 194.03124 182.91502 174.82464 169.25964
#>  [15] 165.79191 164.05578 163.70064 163.70064 157.08380 149.52675 143.22556
#>  [22] 137.99836 133.68818 130.15968 127.29618 124.99722 123.17637 121.75931
#>  [29] 120.68222 119.89033 119.33674 118.98130 118.78970 118.73267 101.22286
#>  [36] 100.93991  92.20846  91.03051  91.03051  91.03050  91.03050  91.03050
#>  [43]  91.03050  90.98642  89.15642  88.20927  87.97838  87.97837  87.97837
#>  [50]  87.97837  87.97837  87.97837  85.68452  85.68452  85.68451  85.66678
#>  [57]  84.98363  84.98362  84.98361  84.98361  84.38399  84.38398  84.38397
#>  [64]  84.38396  84.38396  84.38395  84.38395  84.38394  84.38394  83.55460
#>  [71]  83.55458  83.55457  83.55456  83.55455  83.55454  83.55452  83.55451
#>  [78]  83.55378  83.38288  83.38286  83.38285  83.38283  83.38282  83.38280
#>  [85]  83.38279  83.38278  83.38277  83.38275  83.35807  83.35805  83.35804
#>  [92]  83.35803  83.35801  83.35800  83.35799  83.35798  83.35797  83.35796
#>  [99]  83.35795  83.35794
#> 
#> $linear.predictors
#>      4.8744   4.5459   4.2395   3.9538   3.6873   3.4388   3.2070   2.9909
#> 1  19.48095 19.84489 20.18430 20.50083 20.79603 21.07134 21.32809 21.56753
#> 2  19.48095 19.72206 19.94692 20.15663 20.35220 20.53459 20.70469 20.86333
#> 3  19.48095 19.55829 19.63042 19.69769 19.76043 19.81894 19.87350 19.92439
#> 4  19.48095 19.44029 19.40236 19.36699 19.33400 19.30324 19.27455 19.24780
#> 5  19.48095 19.57034 19.65370 19.73144 19.80394 19.87156 19.93462 19.99342
#> 6  19.48095 19.44992 19.42098 19.39399 19.36881 19.34534 19.32344 19.30303
#> 7  19.48095 19.44992 19.42098 19.39399 19.36881 19.34534 19.32344 19.30303
#> 8  19.48095 19.14647 18.83452 18.54360 18.27229 18.01927 17.78329 17.56322
#> 9  19.48095 19.31023 19.15102 19.00254 18.86407 18.73492 18.61449 18.50217
#> 10 19.48095 18.57809 17.73609 16.95083 16.21849 15.53551 14.89857 14.30455
#> 11 19.48095 18.49428 17.57411 16.71596 15.91564 15.16927 14.47319 13.82403
#> 12 19.48095 18.53234 17.64765 16.82260 16.05314 15.33555 14.66632 14.04219
#> 13 19.48095 20.22300 20.91504 21.56043 22.16233 22.72366 23.24716 23.73538
#> 14 19.48095 19.91955 20.32858 20.71005 21.06581 21.39759 21.70701 21.99558
#> 15 19.48095 19.41139 19.34651 19.28600 19.22957 19.17695 19.12787 19.08210
#> 16 19.48095 19.25484 19.04397 18.84731 18.66391 18.49286 18.33335 18.18458
#> 17 19.48095 20.07609 20.63112 21.14874 21.63148 22.08168 22.50154 22.89310
#> 18 19.48095 20.37810 21.21478 21.99507 22.72278 23.40143 24.03435 24.62461
#> 19 19.48095 19.57997 19.67231 19.75844 19.83875 19.91365 19.98351 20.04866
#> 20 19.48095 19.77264 20.04467 20.29836 20.53496 20.75560 20.96138 21.15329
#> 21 19.48095 19.38730 19.29996 19.21851 19.14255 19.07170 19.00564 18.94402
#>      2.7893   2.6013   2.4260   2.2625   2.1100   1.9678   1.8352    1.7115
#> 1  21.79084 21.99910 22.19332 22.37445 22.54338 22.70092 22.84784 22.984858
#> 2  21.01127 21.14924 21.27792 21.39792 21.50983 21.61420 21.71154 21.802315
#> 3  19.97184 20.01610 20.05738 20.09587 20.13177 20.16525 20.19647 20.225590
#> 4  19.22284 19.19957 19.17787 19.15763 19.13875 19.12115 19.10473 19.089421
#> 5  20.04827 20.09942 20.14712 20.19161 20.23310 20.27179 20.30787 20.341526
#> 6  19.28399 19.26623 19.24967 19.23422 19.21982 19.20638 19.19385 19.182170
#> 7  19.28399 19.26623 19.24967 19.23422 19.21982 19.20638 19.19385 19.182170
#> 8  17.35799 17.16658 16.98808 16.82160 16.66635 16.52156 16.38652 16.260592
#> 9  18.39741 18.29972 18.20862 18.12365 18.04441 17.97051 17.90159 17.837316
#> 10 13.75056 13.23391 12.75208 12.30273 11.88366 11.49284 11.12835 10.788430
#> 11 13.21862 12.65401 12.12745 11.63639 11.17842 10.75131 10.35299  9.981518
#> 12 13.46013 12.91730 12.41105 11.93892 11.49861 11.08798 10.70502 10.347874
#> 13 24.19070 24.61532 25.01133 25.38065 25.72508 26.04629 26.34586 26.625237
#> 14 22.26470 22.51568 22.74974 22.96803 23.17161 23.36147 23.53853 23.703659
#> 15 19.03941 18.99961 18.96248 18.92786 18.89557 18.86545 18.83737 18.811176
#> 16 18.04584 17.91645 17.79579 17.68325 17.57830 17.48042 17.38914 17.304012
#> 17 23.25827 23.59883 23.91644 24.21264 24.48888 24.74650 24.98676 25.210823
#> 18 25.17510 25.68848 26.16726 26.61377 27.03019 27.41854 27.78072 28.118488
#> 19 20.10941 20.16607 20.21892 20.26820 20.31416 20.35702 20.39699 20.434275
#> 20 21.33227 21.49918 21.65485 21.80002 21.93541 22.06167 22.17943 22.289245
#> 21 18.88656 18.83297 18.78299 18.73638 18.69291 18.65237 18.61456 18.579304
#>       1.5961    1.4886    1.3882    1.2947    1.2074    1.1260    1.0502
#> 1  23.080901 23.080901 22.997523 22.892919 22.795364 22.704385 22.619537
#> 2  21.865944 21.865944 21.821604 21.765975 21.714096 21.665714 21.620592
#> 3  20.246001 20.246001 20.368155 20.521409 20.664334 20.797626 20.921934
#> 4  19.078689 19.078689 19.254899 19.475969 19.682140 19.874416 20.053733
#> 5  20.365114 20.365114 20.525497 20.726710 20.914363 21.089368 21.252579
#> 6  19.173980 19.173980 19.218817 19.275069 19.327530 19.376456 19.422084
#> 7  19.173980 19.173980 19.259853 19.367588 19.468062 19.561765 19.649152
#> 8  16.172322 16.172322 16.157427 16.138741 16.121314 16.105061 16.089904
#> 9  17.792264 17.792264 17.790065 17.787306 17.784733 17.782333 17.780095
#> 10 10.550168 10.550168 10.578354 10.613716 10.646694 10.677450 10.706134
#> 11  9.721138  9.721138  9.738885  9.761149  9.781913  9.801278  9.819337
#> 12 10.097537 10.097537 10.087697 10.075353 10.063841 10.053105 10.043092
#> 13 26.821062 26.821062 26.970687 27.158403 27.333468 27.496735 27.648998
#> 14 23.819404 23.819404 23.978373 24.177814 24.363814 24.537278 24.699050
#> 15 18.792817 18.792817 18.740084 18.673925 18.612225 18.554684 18.501020
#> 16 17.244342 17.244342 17.204860 17.155325 17.109129 17.066046 17.025867
#> 17 25.367879 25.367879 25.299527 25.213774 25.133801 25.059218 24.989661
#> 18 28.355243 28.355243 28.298757 28.227891 28.161800 28.100164 28.042681
#> 19 20.460405 20.460405 20.244568 19.973782 19.721245 19.485729 19.266085
#> 20 22.366220 22.366220 22.217619 22.031186 21.857318 21.695168 21.543947
#> 21 18.554590 18.554590 18.346749 18.085995 17.842814 17.616023 17.404517
#>       0.9794    0.9134    0.8518    0.7944    0.7409    0.6909    0.6444
#> 1  22.540408 22.466611 22.397789 22.333605 22.273746 22.217922 22.165860
#> 2  21.578511 21.539266 21.502667 21.468534 21.436701 21.407014 21.379328
#> 3  21.037865 21.145983 21.246813 21.340848 21.428545 21.510332 21.586607
#> 4  20.220965 20.376926 20.522375 20.658022 20.784527 20.902505 21.012533
#> 5  21.404789 21.546742 21.679127 21.802590 21.917732 22.025113 22.125258
#> 6  19.464636 19.504321 19.541331 19.575847 19.608037 19.638057 19.666054
#> 7  19.730650 19.806655 19.877538 19.943643 20.005293 20.062788 20.116408
#> 8  16.075768 16.062585 16.050291 16.038825 16.028132 16.018160 16.008859
#> 9  17.778008 17.776062 17.774246 17.772553 17.770974 17.769502 17.768129
#> 10 10.732884 10.757831 10.781097 10.802795 10.823030 10.841902 10.859501
#> 11  9.836179  9.851887  9.866535  9.880196  9.892937  9.904819  9.915900
#> 12 10.033754 10.025046 10.016924 10.009350 10.002286  9.995698  9.989555
#> 13 27.790998 27.923429 28.046934 28.162115 28.269533 28.369712 28.463139
#> 14 24.849920 24.990622 25.121841 25.244216 25.358344 25.464779 25.564042
#> 15 18.450974 18.404300 18.360772 18.320177 18.282319 18.247012 18.214085
#> 16 16.988396 16.953451 16.920860 16.890466 16.862121 16.835685 16.811032
#> 17 24.924792 24.864295 24.807876 24.755259 24.706188 24.660424 24.617745
#> 18 27.989073 27.939078 27.892453 27.848969 27.808417 27.770597 27.735327
#> 19 19.061245 18.870211 18.692051 18.525899 18.370945 18.226434 18.091663
#> 20 21.402917 21.271393 21.148732 21.034339 20.927655 20.828161 20.735373
#> 21 17.207265 17.023308 16.851749 16.691752 16.542538 16.403382 16.273603
#>       0.6009    0.5604    0.5227    0.4874   0.4546   0.4239   0.3954   0.3687
#> 1  22.117307 22.072026 22.029797 21.990983 22.65806 22.65751 22.62566 22.60872
#> 2  21.353507 21.329427 21.306970 21.286329 22.59999 22.59948 22.56274 22.54305
#> 3  21.657741 21.724081 21.785949 21.842816 21.41926 21.42445 21.66864 21.79715
#> 4  21.115144 21.210840 21.300087 21.382117 21.73604 21.74027 21.93229 22.03315
#> 5  22.218653 22.305753 22.386984 22.461646 22.37178 22.36519 22.06174 21.90222
#> 6  19.692164 19.716514 19.739223 19.760096 18.85182 18.83600 18.11405 17.73473
#> 7  20.166415 20.213051 20.256543 20.296520 19.77569 19.75741 18.91774 18.47642
#> 8  16.000186 15.992097 15.984553 15.977620 15.19219 15.19507 15.33655 15.41116
#> 9  17.766848 17.765654 17.764540 17.763516 16.72728 16.72624 16.68850 16.66896
#> 10 10.875915 10.891222 10.905498 10.918619 11.82638 11.82844 11.91444 11.95940
#> 11  9.926234  9.935872  9.944860  9.953122 10.95200 10.95631 11.14534 11.24444
#> 12  9.983825  9.978482  9.973498  9.968918 10.62121 10.62644 10.86143 10.98479
#> 13 28.550269 28.631528 28.707309 28.776963 30.78820 30.79410 31.04658 31.17875
#> 14 25.656614 25.742947 25.823462 25.897467 25.00387 24.99985 24.82344 24.73094
#> 15 18.183377 18.154738 18.128030 18.103481 16.36117 16.37045 16.81696 17.05222
#> 16 16.788040 16.766598 16.746601 16.728220 15.46533 15.47685 16.02185 16.30874
#> 17 24.577942 24.540821 24.506203 24.474384 24.76004 24.77261 25.35102 25.65504
#> 18 27.702433 27.671757 27.643148 27.616852 27.33098 27.33700 27.61803 27.76587
#> 19 17.965975 17.848758 17.739441 17.638963 17.65498 17.66751 18.24665 18.55113
#> 20 20.648839 20.568136 20.492873 20.423695 20.63847 20.62474 19.98803 19.65322
#> 21 16.152572 16.039697 15.934430 15.837675 16.36527 16.34406 15.35832 14.83992
#>      0.3439   0.3207   0.2991   0.2789   0.2601   0.2426   0.2262   0.2110
#> 1  22.60886 22.60897 22.60905 22.60911 22.60916 22.60922 22.58779 22.56753
#> 2  22.54332 22.54354 22.54370 22.54383 22.54393 22.54446 22.52449 22.50531
#> 3  21.79709 21.79704 21.79700 21.79697 21.79695 21.79423 21.67588 21.56574
#> 4  22.03325 22.03333 22.03338 22.03343 22.03346 22.02837 21.75945 21.50864
#> 5  21.90217 21.90213 21.90210 21.90208 21.90207 21.90403 22.00109 22.09162
#> 6  17.73446 17.73425 17.73409 17.73397 17.73388 17.73788 17.95891 18.16537
#> 7  18.47622 18.47607 18.47595 18.47586 18.47580 18.47967 18.67777 18.86270
#> 8  15.41101 15.41089 15.41080 15.41073 15.41067 15.40940 15.37520 15.34365
#> 9  16.66873 16.66855 16.66842 16.66832 16.66824 16.66613 16.59380 16.52685
#> 10 11.95960 11.95976 11.95988 11.95997 11.96004 11.95973 11.91372 11.87042
#> 11 11.24468 11.24486 11.24500 11.24510 11.24518 11.24593 11.25284 11.25883
#> 12 10.98495 10.98508 10.98518 10.98525 10.98531 10.98795 11.10566 11.21504
#> 13 31.17922 31.17957 31.17985 31.18005 31.18021 31.18184 31.20030 31.21659
#> 14 24.73073 24.73057 24.73045 24.73036 24.73028 24.73109 24.80023 24.86508
#> 15 17.05189 17.05163 17.05144 17.05129 17.05118 17.04649 16.87797 16.72164
#> 16 16.30853 16.30836 16.30824 16.30814 16.30806 16.30688 16.30334 16.30055
#> 17 25.65517 25.65526 25.65533 25.65538 25.65542 25.65983 25.88858 26.10160
#> 18 27.76584 27.76581 27.76579 27.76578 27.76577 27.76292 27.63345 27.51289
#> 19 18.55119 18.55124 18.55127 18.55130 18.55132 18.55587 18.80137 19.03013
#> 20 19.65319 19.65317 19.65316 19.65315 19.65314 19.65041 19.48823 19.33705
#> 21 14.83992 14.83992 14.83993 14.83993 14.83993 14.83767 14.67992 14.53277
#>      0.1968   0.1835   0.1711   0.1596   0.1489   0.1388   0.1295   0.1207
#> 1  22.55142 22.55128 22.55118 22.55111 22.55106 22.55102 22.43311 22.43300
#> 2  22.49005 22.48979 22.48960 22.48946 22.48935 22.48927 22.24964 22.24944
#> 3  21.47833 21.47841 21.47847 21.47851 21.47855 21.47857 21.47567 21.47560
#> 4  21.30946 21.30937 21.30931 21.30928 21.30925 21.30923 21.46268 21.46246
#> 5  22.16345 22.16346 22.16347 22.16348 22.16348 22.16349 22.77771 22.77801
#> 6  18.32912 18.32930 18.32943 18.32952 18.32960 18.32965 17.94897 17.94910
#> 7  19.00931 19.00940 19.00947 19.00952 19.00957 19.00960 18.35387 18.35385
#> 8  15.31864 15.31880 15.31892 15.31900 15.31907 15.31911 15.08561 15.08556
#> 9  16.47364 16.47384 16.47399 16.47410 16.47418 16.47425 15.96907 15.96891
#> 10 11.83606 11.83588 11.83575 11.83565 11.83558 11.83552 12.12526 12.12525
#> 11 11.26367 11.26348 11.26333 11.26323 11.26314 11.26308 11.63424 11.63430
#> 12 11.30205 11.30194 11.30185 11.30178 11.30172 11.30168 11.53822 11.53837
#> 13 31.22965 31.22926 31.22896 31.22875 31.22858 31.22846 30.36785 30.36743
#> 14 24.91655 24.91673 24.91685 24.91695 24.91702 24.91707 25.13454 25.13474
#> 15 16.59759 16.59794 16.59819 16.59839 16.59853 16.59863 16.69278 16.69279
#> 16 16.29854 16.29881 16.29900 16.29914 16.29923 16.29931 16.09877 16.09880
#> 17 26.27106 26.27104 26.27101 26.27098 26.27095 26.27093 26.78294 26.78331
#> 18 27.41720 27.41725 27.41730 27.41733 27.41735 27.41737 28.29822 28.29843
#> 19 19.21208 19.21212 19.21213 19.21213 19.21212 19.21211 18.27652 18.27644
#> 20 19.21668 19.21660 19.21656 19.21654 19.21653 19.21653 19.94701 19.94708
#> 21 14.41546 14.41531 14.41522 14.41516 14.41513 14.41511 14.44733 14.44713
#>      0.1126   0.1050   0.0979   0.0913   0.0852   0.0794   0.0741   0.0691
#> 1  22.43290 22.43259 22.43097 22.43106 22.43115 22.43122 22.25102 22.25089
#> 2  22.24927 22.24906 22.26840 22.26857 22.26872 22.26885 22.05575 22.05553
#> 3  21.47554 21.47320 21.36114 21.36120 21.36126 21.36130 21.56358 21.56358
#> 4  21.46228 21.45730 21.17399 21.17415 21.17429 21.17441 21.32413 21.32403
#> 5  22.77827 22.78026 22.77017 22.76987 22.76961 22.76938 22.73256 22.73280
#> 6  17.94921 17.95331 18.22945 18.22939 18.22934 18.22929 18.39816 18.39821
#> 7  18.35383 18.35754 18.65731 18.65739 18.65746 18.65753 18.72653 18.72641
#> 8  15.08552 15.08468 15.09218 15.09225 15.09231 15.09236 14.95716 14.95704
#> 9  15.96877 15.96718 15.98401 15.98421 15.98437 15.98452 15.73563 15.73534
#> 10 12.12525 12.12456 12.03083 12.03080 12.03077 12.03075 12.01130 12.01134
#> 11 11.63435 11.63465 11.57783 11.57773 11.57764 11.57757 11.62166 11.62180
#> 12 11.53849 11.54077 11.61293 11.61277 11.61264 11.61253 11.69063 11.69082
#> 13 30.36707 30.36750 30.52686 30.52726 30.52760 30.52791 30.52522 30.52486
#> 14 25.13492 25.13616 25.16696 25.16678 25.16663 25.16649 25.03117 25.03125
#> 15 16.69280 16.68930 16.51477 16.51477 16.51477 16.51476 16.43163 16.43161
#> 16 16.09883 16.09840 16.13046 16.13046 16.13046 16.13047 16.07490 16.07489
#> 17 26.78364 26.78789 26.92068 26.92031 26.92000 26.91973 26.67654 26.67684
#> 18 28.29860 28.29623 28.02524 28.02499 28.02478 28.02459 28.39701 28.39740
#> 19 18.27636 18.28062 18.67241 18.67256 18.67269 18.67280 18.95747 18.95747
#> 20 19.94714 19.94444 19.66639 19.66627 19.66617 19.66607 19.54761 19.54769
#> 21 14.44697 14.44436 14.28705 14.28721 14.28735 14.28747 14.39033 14.39020
#>      0.0644   0.0601   0.0560   0.0523   0.0487   0.0455   0.0424   0.0395
#> 1  22.25077 22.25066 22.25056 22.25048 22.25040 22.25034 22.25027 22.30286
#> 2  22.05534 22.05517 22.05502 22.05489 22.05476 22.05465 22.05456 22.19757
#> 3  21.56358 21.56358 21.56358 21.56358 21.56358 21.56358 21.56358 21.45108
#> 4  21.32394 21.32386 21.32379 21.32372 21.32366 21.32360 21.32356 21.52435
#> 5  22.73302 22.73322 22.73339 22.73355 22.73369 22.73382 22.73394 22.67894
#> 6  18.39825 18.39829 18.39833 18.39836 18.39839 18.39842 18.39844 18.61659
#> 7  18.72630 18.72620 18.72611 18.72603 18.72596 18.72590 18.72585 18.76191
#> 8  14.95693 14.95684 14.95676 14.95669 14.95662 14.95656 14.95651 15.39641
#> 9  15.73508 15.73486 15.73465 15.73447 15.73431 15.73417 15.73404 15.76609
#> 10 12.01138 12.01141 12.01144 12.01146 12.01148 12.01150 12.01152 11.71851
#> 11 11.62192 11.62202 11.62212 11.62220 11.62228 11.62234 11.62240 11.63781
#> 12 11.69099 11.69114 11.69128 11.69140 11.69151 11.69161 11.69169 11.86852
#> 13 30.52454 30.52425 30.52399 30.52376 30.52355 30.52336 30.52319 30.44523
#> 14 25.03133 25.03139 25.03146 25.03151 25.03156 25.03161 25.03165 24.70958
#> 15 16.43159 16.43158 16.43156 16.43155 16.43154 16.43153 16.43153 16.46390
#> 16 16.07488 16.07488 16.07487 16.07486 16.07486 16.07485 16.07485 15.70779
#> 17 26.67712 26.67736 26.67758 26.67778 26.67795 26.67811 26.67825 26.70109
#> 18 28.39774 28.39805 28.39833 28.39857 28.39879 28.39898 28.39915 28.50994
#> 19 18.95746 18.95745 18.95744 18.95744 18.95743 18.95742 18.95741 18.89353
#> 20 19.54775 19.54781 19.54786 19.54791 19.54795 19.54799 19.54803 19.65454
#> 21 14.39008 14.38997 14.38988 14.38979 14.38972 14.38965 14.38959 14.09375
#>      0.0369   0.0344   0.0321   0.0299   0.0279   0.0260   0.0243   0.0226
#> 1  22.30286 22.30286 22.30286 22.30285 22.30285 22.30285 22.30285 22.30249
#> 2  22.19760 22.19764 22.19767 22.19770 22.19773 22.19776 22.19779 22.19759
#> 3  21.45105 21.45102 21.45099 21.45096 21.45093 21.45090 21.45088 21.45044
#> 4  21.52443 21.52452 21.52461 21.52469 21.52477 21.52485 21.52492 21.52482
#> 5  22.67891 22.67889 22.67886 22.67884 22.67881 22.67879 22.67877 22.67916
#> 6  18.61670 18.61681 18.61691 18.61702 18.61712 18.61721 18.61731 18.61775
#> 7  18.76192 18.76193 18.76193 18.76194 18.76195 18.76195 18.76196 18.76231
#> 8  15.39659 15.39676 15.39692 15.39708 15.39723 15.39739 15.39753 15.39767
#> 9  15.76605 15.76601 15.76598 15.76594 15.76590 15.76587 15.76584 15.76562
#> 10 11.71838 11.71825 11.71813 11.71801 11.71790 11.71778 11.71767 11.71747
#> 11 11.63784 11.63787 11.63790 11.63792 11.63795 11.63797 11.63800 11.63804
#> 12 11.86864 11.86875 11.86886 11.86896 11.86907 11.86917 11.86927 11.86938
#> 13 30.44520 30.44517 30.44514 30.44511 30.44509 30.44506 30.44503 30.44458
#> 14 24.70941 24.70925 24.70909 24.70894 24.70879 24.70864 24.70850 24.70770
#> 15 16.46389 16.46389 16.46388 16.46388 16.46387 16.46387 16.46386 16.46365
#> 16 15.70763 15.70747 15.70731 15.70716 15.70702 15.70687 15.70673 15.70629
#> 17 26.70112 26.70114 26.70116 26.70119 26.70121 26.70123 26.70125 26.70180
#> 18 28.51005 28.51015 28.51026 28.51036 28.51046 28.51055 28.51064 28.51094
#> 19 18.89358 18.89362 18.89367 18.89371 18.89375 18.89379 18.89383 18.89437
#> 20 19.65455 19.65456 19.65457 19.65458 19.65459 19.65460 19.65461 19.65503
#> 21 14.09359 14.09344 14.09330 14.09315 14.09301 14.09288 14.09274 14.09289
#>      0.0211   0.0197   0.0184   0.0171   0.0160   0.0149   0.0139   0.0129
#> 1  22.14029 22.14022 22.14015 22.14008 22.14002 22.13995 22.13989 22.13983
#> 2  22.12271 22.12265 22.12259 22.12253 22.12247 22.12242 22.12236 22.12231
#> 3  21.39710 21.39711 21.39713 21.39714 21.39715 21.39716 21.39717 21.39718
#> 4  21.63722 21.63720 21.63717 21.63715 21.63713 21.63711 21.63709 21.63707
#> 5  22.83471 22.83479 22.83487 22.83495 22.83502 22.83510 22.83517 22.83524
#> 6  18.56993 18.56980 18.56967 18.56955 18.56942 18.56930 18.56919 18.56907
#> 7  18.82725 18.82727 18.82729 18.82731 18.82733 18.82735 18.82737 18.82739
#> 8  15.40050 15.40034 15.40019 15.40003 15.39989 15.39974 15.39960 15.39946
#> 9  15.82991 15.82999 15.83006 15.83013 15.83020 15.83027 15.83034 15.83040
#> 10 11.75020 11.75033 11.75046 11.75059 11.75072 11.75084 11.75096 11.75108
#> 11 11.61886 11.61882 11.61878 11.61874 11.61871 11.61867 11.61864 11.61860
#> 12 11.75065 11.75048 11.75032 11.75015 11.75000 11.74985 11.74970 11.74955
#> 13 30.47064 30.47070 30.47076 30.47081 30.47087 30.47092 30.47097 30.47102
#> 14 24.50534 24.50540 24.50545 24.50551 24.50557 24.50562 24.50568 24.50573
#> 15 16.47756 16.47758 16.47760 16.47761 16.47762 16.47764 16.47765 16.47767
#> 16 15.69911 15.69926 15.69942 15.69956 15.69971 15.69985 15.69999 15.70012
#> 17 26.82753 26.82755 26.82757 26.82758 26.82760 26.82761 26.82763 26.82764
#> 18 28.46023 28.46010 28.45997 28.45984 28.45972 28.45960 28.45948 28.45937
#> 19 18.96100 18.96099 18.96099 18.96098 18.96097 18.96096 18.96096 18.96095
#> 20 19.71167 19.71167 19.71168 19.71169 19.71169 19.71170 19.71170 19.71171
#> 21 14.10760 14.10775 14.10790 14.10805 14.10819 14.10833 14.10847 14.10860
#>      0.0121   0.0113   0.0105   0.0098   0.0091   0.0085   0.0079   0.0074
#> 1  22.13977 22.13972 22.08040 22.08037 22.08034 22.08031 22.08028 22.08026
#> 2  22.12225 22.12220 22.07928 22.07930 22.07933 22.07935 22.07937 22.07939
#> 3  21.39719 21.39721 21.37761 21.37757 21.37753 21.37749 21.37746 21.37742
#> 4  21.63705 21.63703 21.64789 21.64800 21.64811 21.64821 21.64831 21.64840
#> 5  22.83531 22.83537 22.82689 22.82685 22.82681 22.82677 22.82673 22.82669
#> 6  18.56896 18.56885 18.52590 18.52598 18.52606 18.52615 18.52623 18.52631
#> 7  18.82741 18.82743 18.79883 18.79883 18.79882 18.79882 18.79882 18.79881
#> 8  15.39933 15.39919 15.43578 15.43599 15.43619 15.43638 15.43657 15.43675
#> 9  15.83046 15.83052 15.86790 15.86789 15.86789 15.86788 15.86787 15.86787
#> 10 11.75120 11.75131 11.72833 11.72818 11.72804 11.72790 11.72776 11.72763
#> 11 11.61857 11.61854 11.61631 11.61633 11.61635 11.61638 11.61640 11.61642
#> 12 11.74941 11.74927 11.77158 11.77170 11.77181 11.77193 11.77204 11.77215
#> 13 30.47107 30.47112 30.49217 30.49217 30.49217 30.49217 30.49217 30.49217
#> 14 24.50579 24.50584 24.57997 24.57983 24.57969 24.57956 24.57943 24.57931
#> 15 16.47768 16.47769 16.44811 16.44810 16.44808 16.44806 16.44804 16.44803
#> 16 15.70025 15.70038 15.67644 15.67626 15.67609 15.67592 15.67575 15.67559
#> 17 26.82766 26.82767 26.84740 26.84743 26.84745 26.84748 26.84750 26.84752
#> 18 28.45925 28.45914 28.44832 28.44841 28.44849 28.44858 28.44866 28.44874
#> 19 18.96094 18.96093 18.97646 18.97652 18.97658 18.97663 18.97669 18.97674
#> 20 19.71172 19.71172 19.69908 19.69908 19.69907 19.69907 19.69906 19.69906
#> 21 14.10874 14.10887 14.17536 14.17523 14.17510 14.17498 14.17486 14.17474
#>      0.0069   0.0064   0.0060   0.0056   0.0052   0.0049
#> 1  22.08024 22.08022 22.08020 22.08019 22.08018 22.08017
#> 2  22.07941 22.07943 22.07946 22.07948 22.07950 22.07952
#> 3  21.37739 21.37736 21.37733 21.37730 21.37727 21.37724
#> 4  21.64849 21.64858 21.64866 21.64875 21.64882 21.64890
#> 5  22.82665 22.82661 22.82657 22.82653 22.82650 22.82646
#> 6  18.52638 18.52646 18.52654 18.52661 18.52669 18.52676
#> 7  18.79881 18.79880 18.79880 18.79879 18.79878 18.79878
#> 8  15.43693 15.43710 15.43727 15.43743 15.43759 15.43775
#> 9  15.86786 15.86784 15.86783 15.86782 15.86781 15.86779
#> 10 11.72750 11.72738 11.72725 11.72714 11.72702 11.72691
#> 11 11.61644 11.61646 11.61648 11.61650 11.61652 11.61654
#> 12 11.77226 11.77237 11.77248 11.77258 11.77268 11.77279
#> 13 30.49217 30.49216 30.49216 30.49215 30.49215 30.49214
#> 14 24.57919 24.57907 24.57896 24.57886 24.57875 24.57865
#> 15 16.44801 16.44799 16.44798 16.44796 16.44795 16.44793
#> 16 15.67544 15.67528 15.67514 15.67499 15.67485 15.67471
#> 17 26.84754 26.84756 26.84758 26.84760 26.84761 26.84762
#> 18 28.44882 28.44890 28.44898 28.44906 28.44914 28.44921
#> 19 18.97679 18.97684 18.97688 18.97693 18.97697 18.97701
#> 20 19.69905 19.69904 19.69904 19.69903 19.69903 19.69902
#> 21 14.17463 14.17452 14.17441 14.17430 14.17420 14.17410
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 21.4 18.1 24.4 19.2 17.8 16.4 17.3 10.4 10.4 14.7 33.9 21.5 15.5
#> [16] 19.2 26.0 30.4 15.8 19.7 15.0
#> 
#> $X
#>               am        carb        cyl       disp        drat       gear
#>  [1,]  1.2747549  0.49656353 -0.2399041 -0.6916224  0.77413255  0.2936101
#>  [2,]  1.2747549  0.49656353 -0.2399041 -0.6916224  0.77413255  0.2936101
#>  [3,] -0.7844645 -1.24140883 -0.2399041  0.1124897 -0.87365021 -0.9395524
#>  [4,] -0.7844645 -1.24140883 -0.2399041 -0.1582827 -1.51668739 -0.9395524
#>  [5,] -0.7844645 -0.66208471 -1.4994004 -0.8007519  0.35213941  0.2936101
#>  [6,] -0.7844645  0.49656353 -0.2399041 -0.6292626  0.81432238  0.2936101
#>  [7,] -0.7844645  0.49656353 -0.2399041 -0.6292626  0.81432238  0.2936101
#>  [8,] -0.7844645 -0.08276059  1.0195922  0.2585427 -0.89374512 -0.9395524
#>  [9,] -0.7844645 -0.08276059  1.0195922  0.2585427 -0.89374512 -0.9395524
#> [10,] -0.7844645  0.49656353  1.0195922  1.8684080 -1.17507389 -0.9395524
#> [11,] -0.7844645  0.49656353  1.0195922  1.7699453 -1.03440950 -0.9395524
#> [12,] -0.7844645  0.49656353  1.0195922  1.6058408 -0.57222653 -0.9395524
#> [13,]  1.2747549 -1.24140883 -1.4994004 -1.4210669  1.41716973  0.2936101
#> [14,] -0.7844645 -1.24140883 -1.4994004 -1.0190109  0.37223432 -0.9395524
#> [15,] -0.7844645 -0.66208471  1.0195922  0.6048033 -1.51668739 -0.9395524
#> [16,] -0.7844645 -0.66208471  1.0195922  1.2776317 -0.87365021 -0.9395524
#> [17,]  1.2747549 -0.66208471 -1.4994004 -1.0173698  1.83916287  1.5267726
#> [18,]  1.2747549 -0.66208471 -1.4994004 -1.2241415  0.51289870  1.5267726
#> [19,]  1.2747549  0.49656353  1.0195922  0.8755757  1.41716973  1.5267726
#> [20,]  1.2747549  1.65521178 -0.2399041 -0.8147007  0.21147502  1.5267726
#> [21,]  1.2747549  2.81386002  1.0195922  0.4653144  0.05071573  1.5267726
#>                hp        qsec         vs          wt
#>  [1,] -0.64417450 -0.72504680 -0.7844645 -0.73854444
#>  [2,] -0.64417450 -0.38302713 -0.7844645 -0.48929093
#>  [3,] -0.64417450  1.09498644  1.2747549 -0.15695291
#>  [4,] -0.71866604  1.57137097  1.2747549  0.08252595
#>  [5,] -1.35929332  1.43700610  1.2747549 -0.18138953
#>  [6,] -0.45049648  0.39873211  1.2747549  0.06297666
#>  [7,] -0.45049648  0.76518176  1.2747549  0.06297666
#>  [8,]  0.39870712 -0.15094235 -0.7844645  0.67877945
#>  [9,]  0.39870712 -0.02879247 -0.7844645  0.34644144
#> [10,]  0.77116484  0.20329230 -0.7844645  1.83218787
#> [11,]  0.92014793  0.10557240 -0.7844645  2.00226673
#> [12,]  1.14362256 -0.13872737 -0.7844645  1.92504702
#> [13,] -1.31459839  1.37593116  1.2747549 -1.50585428
#> [14,] -0.83785251  1.44311360  1.2747549 -0.89005148
#> [15,] -0.04824214 -0.47463954 -0.7844645  0.14117384
#> [16,]  0.32421558 -0.36470465 -0.7844645  0.45884988
#> [17,] -0.92724236 -0.57846694 -0.7844645 -1.20772752
#> [18,] -0.59947957 -0.45631706  1.2747549 -1.82059793
#> [19,]  1.65016507 -1.92211563 -0.7844645 -0.20093883
#> [20,]  0.32421558 -1.31136623 -0.7844645 -0.59192473
#> [21,]  2.70794499 -1.86104069 -0.7844645  0.19004708
#> attr(,"center")
#>  [1]   0.3809524   3.1428571   6.3809524 244.2904762   3.5147619   3.7619048
#>  [7] 153.2380952  17.6471429   0.3809524   3.3755714
#> attr(,"scale")
#>  [1]   0.4856209   1.7261494   1.5879364 121.8735563   0.4976384   0.8109232
#>  [7]  67.1217123   1.6373327   0.4856209   1.0230548
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 30
#> 
#> $lambda.min
#> [1] 0.6443647
#> 
#> $null.dev
#> [1] 31.55488
#> 
#> $Bias
#> [1] 3.000275
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 7.768596