Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 31.969245 30.893068 28.372905 25.002613 21.905869 19.435342 17.494448
#>   [8] 16.000193 14.881337 14.076802 13.534285 13.209060 13.049099 12.888335
#>  [15] 12.748655 12.713571 12.211598 11.607877 11.157657 10.638386 10.240100
#>  [22]  9.838804  8.978379  8.306342  7.886764  7.179335  7.082947  6.960189
#>  [29]  6.835699  6.668471  6.590341  6.532409  6.510519  6.503294  6.530929
#>  [36]  7.433992  7.439309  7.521163  7.754021  7.696076  7.722006  7.866439
#>  [43]  7.491080  6.699758  6.640634  6.704858  6.458886  7.746889  7.625107
#>  [50]  7.622819  7.535344  8.926582  8.927219  8.913262  9.135400 10.323716
#>  [57] 10.324594 10.592389 10.358406 10.152352 10.634476 10.635308 15.246520
#>  [64] 15.248618 15.250651 14.919396 14.857630 14.859175 14.860677 14.862138
#>  [71] 14.863559 14.864941 14.800173 14.654031 14.655160 14.656269 14.657358
#>  [78] 14.658426 14.659474 14.660501 14.661508 14.662494 14.663460 14.664405
#>  [85] 14.665330 14.666234 14.667118 14.667982 14.668827 14.669651 14.670457
#>  [92] 14.671243 14.672010 14.672758 14.627705 14.578782 14.579104 14.534642
#>  [99] 14.534829 14.535013
#> 
#> $cvse
#>   [1] 11.250213 11.301557 10.891826 10.000724  8.964941  8.088490  7.332646
#>   [8]  6.668832  6.077279  5.545573  5.067289  4.640784  4.269059  3.965825
#>  [15]  3.728475  3.548758  3.312148  3.105173  2.956585  2.838154  2.766215
#>  [22]  2.697913  2.615537  2.564643  2.517889  2.104718  2.068823  2.025493
#>  [29]  1.986754  1.956838  1.926309  1.897265  1.887479  1.885346  1.889779
#>  [36]  2.026030  2.000728  2.015269  1.951042  1.956597  1.948330  1.953874
#>  [43]  1.949846  1.928149  1.773992  1.765534  1.531260  1.837163  1.852053
#>  [50]  1.852067  1.860367  2.288977  2.289156  2.291207  2.243788  2.216118
#>  [57]  2.216266  2.233590  2.192306  2.218113  2.183330  2.183406  5.481921
#>  [64]  5.483041  5.484129  5.505020  5.511276  5.512301  5.513297  5.514264
#>  [71]  5.515203  5.516114  5.518695  5.507731  5.508561  5.509367  5.510149
#>  [78]  5.510908  5.511644  5.512358  5.513051  5.513724  5.514376  5.515008
#>  [85]  5.515621  5.516216  5.516792  5.517351  5.517892  5.518417  5.518926
#>  [92]  5.519419  5.519896  5.520359  5.521190  5.512347  5.512723  5.509060
#>  [99]  5.509406  5.509739
#> 
#> $fold
#>  [1]  1 10 10  8  5  7  9  3  5  6  2  3  7  1  8  4  2  1  6  4  9
#> 
#> $lambda
#>   [1] 4.807342197 4.483343422 4.181181081 3.899383470 3.636578075 3.391484884
#>   [7] 3.162910154 2.949740595 2.750937952 2.565533941 2.392625540 2.231370586
#>  [13] 2.080983677 1.940732342 1.809933477 1.687950019 1.574187837 1.468092845
#>  [19] 1.369148301 1.276872288 1.190815369 1.110558399 1.035710480 0.965907060
#>  [25] 0.900808157 0.840096702 0.783476996 0.730673268 0.681428336 0.635502346
#>  [31] 0.592671615 0.552727532 0.515475546 0.480734220 0.448334342 0.418118108
#>  [37] 0.389938347 0.363657808 0.339148488 0.316291015 0.294974060 0.275093795
#>  [43] 0.256553394 0.239262554 0.223137059 0.208098368 0.194073234 0.180993348
#>  [49] 0.168795002 0.157418784 0.146809285 0.136914830 0.127687229 0.119081537
#>  [55] 0.111055840 0.103571048 0.096590706 0.090080816 0.084009670 0.078347700
#>  [61] 0.073067327 0.068142834 0.063550235 0.059267162 0.055272753 0.051547555
#>  [67] 0.048073422 0.044833434 0.041811811 0.038993835 0.036365781 0.033914849
#>  [73] 0.031629102 0.029497406 0.027509380 0.025655339 0.023926255 0.022313706
#>  [79] 0.020809837 0.019407323 0.018099335 0.016879500 0.015741878 0.014680928
#>  [85] 0.013691483 0.012768723 0.011908154 0.011105584 0.010357105 0.009659071
#>  [91] 0.009008082 0.008400967 0.007834770 0.007306733 0.006814283 0.006355023
#>  [97] 0.005926716 0.005527275 0.005154755 0.004807342
#> 
#> $fit
#> $beta
#>               4.8073     4.4833    4.1812    3.8994    3.6366   3.3915
#> (Intercept) 19.90476 21.7503959 23.471640 25.076879 26.573929 27.97008
#> am           0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> carb         0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> cyl          0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> disp         0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> drat         0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> gear         0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> hp           0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> qsec         0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> vs           0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> wt           0.00000 -0.5835426 -1.127756 -1.635292 -2.108621 -2.55005
#>                3.1629   2.9497    2.7509    2.5655   2.3926   2.2314    2.0810
#> (Intercept) 29.272142 30.48645 31.618910 32.675050 33.66001 34.57859 35.435253
#> am           0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.000000
#> carb         0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.000000
#> cyl          0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.000000
#> disp         0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.000000
#> drat         0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.000000
#> gear         0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.000000
#> hp           0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.000000
#> qsec         0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.000000
#> vs           0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.000000
#> wt          -2.961728 -3.34566 -3.703716 -4.037641 -4.34906 -4.63949 -4.910347
#>                 1.9407     1.8099     1.6880     1.5742     1.4681     1.3691
#> (Intercept) 34.7774946 33.8537255 32.9933313 32.1909250 30.4031441 28.5294108
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0722231  0.1549654  0.2320759  0.3039893  0.4014762  0.4984412
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.1050283 -5.2742496 -5.4321102 -5.5793316 -5.5575763 -5.5057350
#>                 1.2769     1.1908     1.1106     1.0357    0.9659     0.9008
#> (Intercept) 26.7818360 25.1520418 23.6320903 22.2145784 20.892602 19.6597225
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> qsec         0.5888775  0.6732187  0.7518756  0.8252313  0.893643  0.9574441
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> wt          -5.4573843 -5.4122922 -5.3702391 -5.3310203 -5.294445 -5.2603342
#>                0.8401    0.7835    0.7307    0.6814    0.6355    0.5927
#> (Intercept) 18.509935 17.437639 16.437613 15.504984 15.321275 15.321220
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         1.016945  1.072436  1.124187  1.172450  1.181957  1.181960
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -5.228523 -5.198855 -5.171187 -5.145384 -5.140301 -5.140299
#>                0.5527    0.5155    0.4807    0.4483    0.4181    0.3899
#> (Intercept) 15.321216 15.321216 15.321216 15.321216 15.321216 15.321216
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         1.181960  1.181960  1.181960  1.181960  1.181960  1.181960
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -5.140299 -5.140299 -5.140299 -5.140299 -5.140299 -5.140299
#>                0.3637    0.3391    0.3163     0.2950     0.2751     0.2566
#> (Intercept) 15.321216 15.321216 15.321216 13.7550056 -11.841842 -11.847983
#> am           0.000000  0.000000  0.000000  0.0000000   0.000000   0.000000
#> carb         0.000000  0.000000  0.000000  0.0000000   0.000000   0.000000
#> cyl          0.000000  0.000000  0.000000  0.0000000   0.000000   0.000000
#> disp         0.000000  0.000000  0.000000  0.0000000   0.000000   0.000000
#> drat         0.000000  0.000000  0.000000  0.2508505   2.466181   2.466658
#> gear         0.000000  0.000000  0.000000  0.0000000   0.000000   0.000000
#> hp           0.000000  0.000000  0.000000  0.0000000   0.000000   0.000000
#> qsec         1.181960  1.181960  1.181960  1.1967181   2.237535   2.237784
#> vs           0.000000  0.000000  0.000000  0.0000000  -4.072782  -4.073548
#> wt          -5.140299 -5.140299 -5.140299 -5.0181170  -4.682049  -4.681930
#>                 0.2393        0.2231      0.2081      0.1941       0.1810
#> (Intercept) -11.852994 -11.885958646 -13.7089746 -15.4068905 -30.86169047
#> am            0.000000   0.000000000   0.0000000   0.0000000   0.00000000
#> carb          0.000000   0.000000000   0.0000000   0.0000000   0.00000000
#> cyl           0.000000   0.000000000   0.0000000   0.0000000   0.00000000
#> disp          0.000000   0.000000000   0.0000000   0.0000000   0.01443497
#> drat          2.467047   2.466687864   2.3804283   2.2996787   2.39866034
#> gear          0.000000   0.003212422   0.2210181   0.4240982   1.51297197
#> hp            0.000000   0.000000000   0.0000000   0.0000000   0.00000000
#> qsec          2.237987   2.239179234   2.3129645   2.3817586   3.03174046
#> vs           -4.074173  -4.077105367  -4.2471906  -4.4058835  -5.04931166
#> wt           -4.681833  -4.681023599  -4.6494685  -4.6202271  -5.68899989
#>                   0.1688       0.1574       0.1468       0.1369       0.1277
#> (Intercept) -40.25599783 -40.26066334 -40.26506789 -40.26922607 -40.27315166
#> am            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> carb          0.71299774   0.71318543   0.71336262   0.71352990   0.71368783
#> cyl           0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> disp          0.02269274   0.02269729   0.02270158   0.02270563   0.02270946
#> drat          2.51151062   2.51150375   2.51149727   2.51149115   2.51148537
#> gear          0.92412889   0.92412836   0.92412786   0.92412739   0.92412695
#> hp            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> qsec          3.69315474   3.69344346   3.69371602   3.69397334   3.69421627
#> vs           -5.26917865  -5.26941645  -5.26964095  -5.26985289  -5.27005298
#> wt           -7.07318939  -7.07378126  -7.07434003  -7.07486754  -7.07536554
#>                   0.1191       0.1111      0.1036       0.0966       0.0901
#> (Intercept) -40.27685768 -40.28035640 -40.2836594 -40.28677769 -40.28972154
#> am            0.00000000   0.00000000   0.0000000   0.00000000   0.00000000
#> carb          0.71383692   0.71397767   0.7141105   0.71423599   0.71435442
#> cyl           0.00000000   0.00000000   0.0000000   0.00000000   0.00000000
#> disp          0.02271307   0.02271648   0.0227197   0.02272274   0.02272561
#> drat          2.51147992   2.51147477   2.5114699   2.51146532   2.51146099
#> gear          0.92412653   0.92412614   0.9241258   0.92412541   0.92412508
#> hp            0.00000000   0.00000000   0.0000000   0.00000000   0.00000000
#> qsec          3.69444561   3.69466212   3.6948665   3.69505949   3.69524166
#> vs           -5.27024187  -5.27042020  -5.2705886  -5.27074749  -5.27089754
#> wt           -7.07583569  -7.07627954  -7.0766986  -7.07709415  -7.07746761
#>                   0.0840       0.0783       0.0731       0.0681       0.0636
#> (Intercept) -40.29250073 -40.29512446 -40.29760143 -43.81236955 -43.81814638
#> am            0.00000000   0.00000000   0.00000000  -2.01384948  -2.01417193
#> carb          0.71446622   0.71457178   0.71467142   0.48752313   0.48735642
#> cyl           0.00000000   0.00000000   0.00000000  -1.13116005  -1.13111033
#> disp          0.02272832   0.02273088   0.02273329   0.02021417   0.02020994
#> drat          2.51145690   2.51145304   2.51144939   2.95548557   2.95580869
#> gear          0.92412477   0.92412447   0.92412419   1.17084146   1.17103470
#> hp            0.00000000   0.00000000   0.00000000   0.02524924   0.02525616
#> qsec          3.69541364   3.69557601   3.69572929   4.21621351   4.21644535
#> vs           -5.27103920  -5.27117293  -5.27129918  -8.32323493  -8.32403015
#> wt           -7.07782018  -7.07815303  -7.07846726  -7.48672576  -7.48659088
#>                   0.0593       0.0553       0.0515       0.0481       0.0448
#> (Intercept) -43.82369498 -43.82902427 -43.83414280 -43.83905882 -43.84378024
#> am           -2.01448134  -2.01477822  -2.01506307  -2.01533639  -2.01559864
#> carb          0.48719621   0.48704224   0.48689429   0.48675211   0.48661548
#> cyl          -1.13106230  -1.13101591  -1.13097110  -1.13092783  -1.13088604
#> disp          0.02020588   0.02020198   0.02019824   0.02019463   0.02019117
#> drat          2.95611906   2.95641718   2.95670352   2.95697854   2.95724268
#> gear          1.17122035   1.17139870   1.17157004   1.17173463   1.17189275
#> hp            0.02526281   0.02526919   0.02527531   0.02528120   0.02528684
#> qsec          4.21666791   4.21688156   4.21708665   4.21728353   4.21747251
#> vs           -8.32479347  -8.32552614  -8.32622939  -8.32690440  -8.32755228
#> wt           -7.48646103  -7.48633603  -7.48621569  -7.48609986  -7.48598837
#>                   0.0418       0.0390       0.0364       0.0339       0.0316
#> (Intercept) -43.84831468 -43.85266946 -43.85685162 -43.86086793 -43.86472489
#> am           -2.01585026  -2.01609168  -2.01632330  -2.01654554  -2.01675876
#> carb          0.48648420   0.48635805   0.48623685   0.48612038   0.48600849
#> cyl          -1.13084569  -1.13080673  -1.13076912  -1.13073282  -1.13069778
#> disp          0.02018785   0.02018465   0.02018158   0.02017863   0.02017579
#> drat          2.95749637   2.95774002   2.95797402   2.95819875   2.95841457
#> gear          1.17204463   1.17219053   1.17233067   1.17246528   1.17259458
#> hp            0.02529227   0.02529747   0.02530247   0.02530727   0.02531187
#> qsec          4.21765392   4.21782805   4.21799520   4.21815564   4.21830963
#> vs           -8.32817412  -8.32877095  -8.32934377  -8.32989355  -8.33042120
#> wt           -7.48588106  -7.48577777  -7.48567837  -7.48558271  -7.48549064
#>                   0.0295       0.0275       0.0257       0.0239       0.0223
#> (Intercept) -43.86842875 -43.87198553 -43.87540101 -43.87868073 -43.88183004
#> am           -2.01696332  -2.01715958  -2.01734787  -2.01752852  -2.01770182
#> carb          0.48590098   0.48579768   0.48569845   0.48560311   0.48551152
#> cyl          -1.13066396  -1.13063132  -1.13059982  -1.13056943  -1.13054011
#> disp          0.02017307   0.02017045   0.02016793   0.02016552   0.02016319
#> drat          2.95862183   2.95882087   2.95901201   2.95919556   2.95937181
#> gear          1.17271878   1.17283806   1.17295263   1.17306267   1.17316836
#> hp            0.02531629   0.02532054   0.02532461   0.02532853   0.02533228
#> qsec          4.21845745   4.21859932   4.21873549   4.21886619   4.21899163
#> vs           -8.33092760  -8.33141361  -8.33188003  -8.33232765  -8.33275723
#> wt           -7.48540205  -7.48531681  -7.48523478  -7.48515586  -7.48507993
#>                   0.0208       0.0194       0.0181       0.0169       0.0157
#> (Intercept) -43.88485407 -43.88775775 -43.89054583 -43.89322285 -43.89579320
#> am           -2.01786808  -2.01802759  -2.01818060  -2.01832740  -2.01846822
#> carb          0.48542352   0.48533899   0.48525779   0.48517979   0.48510486
#> cyl          -1.13051183  -1.13048454  -1.13045822  -1.13043283  -1.13040835
#> disp          0.02016096   0.02015882   0.02015676   0.02015479   0.02015289
#> drat          2.95954106   2.95970357   2.95985962   2.96000947   2.96015334
#> gear          1.17326986   1.17336734   1.17346095   1.17355086   1.17363720
#> hp            0.02533589   0.02533935   0.02534267   0.02534585   0.02534891
#> qsec          4.21911202   4.21922757   4.21933846   4.21944489   4.21954704
#> vs           -8.33316948  -8.33356510  -8.33394475  -8.33430908  -8.33465869
#> wt           -7.48500687  -7.48493658  -7.48486897  -7.48480392  -7.48474135
#>                   0.0147       0.0137       0.0128       0.0119       0.0111
#> (Intercept) -43.89826108 -43.90063054 -43.90290547 -43.90508960 -43.90718652
#> am           -2.01860331  -2.01873291  -2.01885722  -2.01897648  -2.01909088
#> carb          0.48503288   0.48496375   0.48489734   0.48483356   0.48477229
#> cyl          -1.13038474  -1.13036197  -1.13034001  -1.13031884  -1.13029844
#> disp          0.02015106   0.02014931   0.02014762   0.02014601   0.02014445
#> drat          2.96029148   2.96042412   2.96055147   2.96067374   2.96079113
#> gear          1.17372011   1.17379973   1.17387619   1.17394961   1.17402012
#> hp            0.02535185   0.02535467   0.02535738   0.02535997   0.02536247
#> qsec          4.21964506   4.21973914   4.21982942   4.21991606   4.21999921
#> vs           -8.33499418  -8.33531612  -8.33562504  -8.33592147  -8.33620592
#> wt           -7.48468116  -7.48462327  -7.48456758  -7.48451402  -7.48446251
#>                   0.0104       0.0097       0.0090       0.0084       0.0078
#> (Intercept) -43.90919968 -43.91113239 -43.91298783 -43.91476908 -43.91647906
#> am           -2.01920061  -2.01930588  -2.01940685  -2.01950370  -2.01959661
#> carb          0.48471345   0.48465693   0.48460265   0.48455052   0.48450045
#> cyl          -1.13027876  -1.13025980  -1.13024152  -1.13022390  -1.13020692
#> disp          0.02014296   0.02014153   0.02014015   0.02013883   0.02013756
#> drat          2.96090383   2.96101204   2.96111592   2.96121565   2.96131139
#> gear          1.17408782   1.17415282   1.17421524   1.17427517   1.17433271
#> hp            0.02536486   0.02536715   0.02536936   0.02537147   0.02537350
#> qsec          4.22007899   4.22015556   4.22022904   4.22029954   4.22036720
#> vs           -8.33647885  -8.33674074  -8.33699203  -8.33723315  -8.33746449
#> wt           -7.48441297  -7.48436532  -7.48431950  -7.48427544  -7.48423307
#>                   0.0073       0.0068       0.0064       0.0059       0.0055
#> (Intercept) -43.91812060 -43.91969642 -43.92120912 -43.92266121 -43.92405510
#> am           -2.01968572  -2.01977120  -2.01985319  -2.01993184  -2.02000727
#> carb          0.48445237   0.48440619   0.48436185   0.48431926   0.48427837
#> cyl          -1.13019055  -1.13017478  -1.13015958  -1.13014494  -1.13013083
#> disp          0.02013634   0.02013517   0.02013404   0.02013296   0.02013193
#> drat          2.96140330   2.96149153   2.96157624   2.96165755   2.96173560
#> gear          1.17438796   1.17444101   1.17449194   1.17454084   1.17458778
#> hp            0.02537545   0.02537732   0.02537911   0.02538083   0.02538249
#> qsec          4.22043212   4.22049442   4.22055420   4.22061156   4.22066659
#> vs           -8.33768647  -8.33789945  -8.33810379  -8.33829985  -8.33848796
#> wt           -7.48419233  -7.48415315  -7.48411549  -7.48407927  -7.48404445
#>                   0.0052       0.0048
#> (Intercept) -43.92539309 -43.92667741
#> am           -2.02007962  -2.02014901
#> carb          0.48423910   0.48420139
#> cyl          -1.13011724  -1.13010415
#> disp          0.02013093   0.02012997
#> drat          2.96181053   2.96188245
#> gear          1.17463286   1.17467613
#> hp            0.02538407   0.02538559
#> qsec          4.22071940   4.22077007
#> vs           -8.33866843  -8.33884158
#> wt           -7.48401097  -7.48397878
#> 
#> $iter
#>   [1]   0   6   4   3   3   2   2   2   2   2   2   2   2   6   5   5   5   4
#>  [19]   4   4   4   4   4   4   4   4   4   4   4   3   1   1   1   1   1   1
#>  [37]   1   1   1   1  26  66   1   1   4  35  34  87 103   1   1   1   1   1
#>  [55]   1   1   1   1   1   1   1 244   1   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.807342197 4.483343422 4.181181081 3.899383470 3.636578075 3.391484884
#>   [7] 3.162910154 2.949740595 2.750937952 2.565533941 2.392625540 2.231370586
#>  [13] 2.080983677 1.940732342 1.809933477 1.687950019 1.574187837 1.468092845
#>  [19] 1.369148301 1.276872288 1.190815369 1.110558399 1.035710480 0.965907060
#>  [25] 0.900808157 0.840096702 0.783476996 0.730673268 0.681428336 0.635502346
#>  [31] 0.592671615 0.552727532 0.515475546 0.480734220 0.448334342 0.418118108
#>  [37] 0.389938347 0.363657808 0.339148488 0.316291015 0.294974060 0.275093795
#>  [43] 0.256553394 0.239262554 0.223137059 0.208098368 0.194073234 0.180993348
#>  [49] 0.168795002 0.157418784 0.146809285 0.136914830 0.127687229 0.119081537
#>  [55] 0.111055840 0.103571048 0.096590706 0.090080816 0.084009670 0.078347700
#>  [61] 0.073067327 0.068142834 0.063550235 0.059267162 0.055272753 0.051547555
#>  [67] 0.048073422 0.044833434 0.041811811 0.038993835 0.036365781 0.033914849
#>  [73] 0.031629102 0.029497406 0.027509380 0.025655339 0.023926255 0.022313706
#>  [79] 0.020809837 0.019407323 0.018099335 0.016879500 0.015741878 0.014680928
#>  [85] 0.013691483 0.012768723 0.011908154 0.011105584 0.010357105 0.009659071
#>  [91] 0.009008082 0.008400967 0.007834770 0.007306733 0.006814283 0.006355023
#>  [97] 0.005926716 0.005527275 0.005154755 0.004807342
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 40
#> 
#> $loss
#>   [1] 645.58952 552.42250 474.47444 409.55520 355.77394 311.49916 275.32403
#>   [8] 246.03636 222.59240 204.09427 189.77023 178.95768 171.08831 156.60890
#>  [15] 143.30733 132.80623 124.66301 115.45321 107.37858 100.81162  95.52573
#>  [22]  91.32535  88.04235  85.53227  83.67117  82.35281  81.48626  80.99380
#>  [29]  80.80909  80.80395  80.80395  80.80395  80.80395  80.80395  80.80395
#>  [36]  80.80395  80.80395  80.80395  80.80395  80.80395  79.27347  61.27861
#>  [43]  61.27860  61.27859  61.25799  60.06630  59.31931  51.90883  47.33961
#>  [50]  47.33959  47.33957  47.33956  47.33955  47.33953  47.33952  47.33951
#>  [57]  47.33951  47.33950  47.33949  47.33949  47.33948  41.72497  41.72495
#>  [64]  41.72493  41.72491  41.72489  41.72488  41.72486  41.72485  41.72484
#>  [71]  41.72483  41.72482  41.72481  41.72480  41.72479  41.72479  41.72478
#>  [78]  41.72477  41.72477  41.72476  41.72476  41.72475  41.72475  41.72474
#>  [85]  41.72474  41.72474  41.72473  41.72473  41.72473  41.72473  41.72473
#>  [92]  41.72472  41.72472  41.72472  41.72472  41.72472  41.72472  41.72471
#>  [99]  41.72471  41.72471
#> 
#> $linear.predictors
#>      4.8073   4.4833   4.1812   3.8994   3.6366   3.3915   3.1629   2.9497
#> 1  19.90476 20.22151 20.51692 20.79241 21.04934 21.28895 21.51242 21.72082
#> 2  19.90476 20.07271 20.22934 20.37541 20.51164 20.63869 20.75718 20.86767
#> 3  19.90476 20.39658 20.85525 21.28300 21.68193 22.05397 22.40093 22.72452
#> 4  19.90476 19.87431 19.84590 19.81942 19.79471 19.77167 19.75019 19.73015
#> 5  19.90476 19.73134 19.56960 19.41877 19.27810 19.14691 19.02456 18.91046
#> 6  19.90476 19.66715 19.44555 19.23889 19.04615 18.86641 18.69877 18.54244
#> 7  19.90476 19.88890 19.87410 19.86030 19.84743 19.83542 19.82423 19.81379
#> 8  19.90476 19.74301 19.59216 19.45147 19.32027 19.19791 19.08380 18.97738
#> 9  19.90476 19.74301 19.59216 19.45147 19.32027 19.19791 19.08380 18.97738
#> 10 19.90476 19.37538 18.88167 18.42124 17.99184 17.59138 17.21791 16.86961
#> 11 19.90476 18.58526 17.35469 16.20706 15.13677 14.13861 13.20773 12.33959
#> 12 19.90476 20.80797 21.65031 22.43588 23.16851 23.85175 24.48895 25.08321
#> 13 19.90476 20.67960 21.40221 22.07612 22.70461 23.29074 23.83737 24.34716
#> 14 19.90476 19.74593 19.59780 19.45965 19.33082 19.21066 19.09861 18.99411
#> 15 19.90476 19.50959 19.14106 18.79736 18.47682 18.17789 17.89911 17.63911
#> 16 19.90476 19.50667 19.13542 18.78918 18.46628 18.16514 17.88430 17.62238
#> 17 19.90476 20.62124 21.28943 21.91259 22.49375 23.03574 23.54120 24.01259
#> 18 19.90476 19.90057 19.89665 19.89300 19.88960 19.88643 19.88347 19.88071
#> 19 19.90476 20.13398 20.34776 20.54712 20.73305 20.90645 21.06816 21.21897
#> 20 19.90476 19.66715 19.44555 19.23889 19.04615 18.86641 18.69877 18.54244
#> 21 19.90476 20.12815 20.33648 20.53077 20.71196 20.88095 21.03854 21.18551
#>      2.7509   2.5655   2.3926    2.2314    2.0810    1.9407   1.8099    1.6880
#> 1  21.91517 22.09643 22.26547 22.423121 22.570145 22.591113 22.58592 22.581171
#> 2  20.97073 21.06683 21.15646 21.240051 21.318006 21.329775 21.32777 21.325946
#> 3  23.02629 23.30772 23.57019 23.814968 24.043249 24.277901 24.50137 24.709767
#> 4  19.71146 19.69404 19.67778 19.662625 19.648489 19.768845 19.90954 20.040652
#> 5  18.80405 18.70481 18.61226 18.525950 18.445454 18.574448 18.73822 18.890804
#> 6  18.39664 18.26067 18.13387 18.015606 17.905316 17.696557 17.47931 17.276779
#> 7  19.80406 19.79498 19.78651 19.778612 19.771247 19.936916 20.12818 20.306417
#> 8  18.87813 18.78557 18.69924 18.618739 18.543661 18.537880 18.54617 18.553860
#> 9  18.87813 18.78557 18.69924 18.618739 18.543661 18.581214 18.63915 18.693106
#> 10 16.54479 16.24185 15.95934 15.695861 15.450142 15.256711 15.08393 14.922763
#> 11 11.52995 10.77489 10.07071  9.413991  8.801533  8.374836  8.00768  7.665157
#> 12 25.63741 26.15426 26.63628 27.085809 27.505043 27.870446 28.20577 28.518518
#> 13 24.82259 25.26598 25.67948 26.065121 26.424767 26.847007 27.25929 27.643719
#> 14 18.89665 18.80575 18.72099 18.641937 18.568212 18.491182 18.41758 18.348945
#> 15 17.39664 17.17051 16.95962 16.762943 16.579522 16.287144 15.98862 15.710317
#> 16 17.37812 17.15032 16.93787 16.739746 16.554970 16.380064 16.21640 16.063761
#> 17 24.45222 24.86222 25.24458 25.601172 25.933732 26.264281 26.57690 26.868432
#> 18 19.87813 19.87573 19.87349 19.871402 19.869454 19.641790 19.38135 19.138642
#> 19 21.35962 21.49079 21.61311 21.727198 21.833593 21.756024 21.64602 21.543562
#> 20 18.39664 18.26067 18.13387 18.015606 17.905316 17.607001 17.28715 16.989005
#> 21 21.32258 21.45041 21.56962 21.680803 21.784489 21.928865 22.07367 22.208676
#>      1.5742    1.4681    1.3691    1.2769    1.1908    1.1106    1.0357
#> 1  22.57674 22.450593 22.308727 22.176413 22.053016 21.937936 21.830612
#> 2  21.32424 21.258238 21.183892 21.114551 21.049884 20.989576 20.933331
#> 3  24.90412 24.981040 25.032096 25.079715 25.124124 25.165540 25.204165
#> 4  20.16293 20.340234 20.518169 20.684124 20.838894 20.983233 21.117844
#> 5  19.03310 19.291780 19.558048 19.806389 20.037993 20.253988 20.455425
#> 6  17.08790 16.921980 16.769245 16.626794 16.493943 16.370046 16.254499
#> 7  20.47264 20.704000 20.934940 21.150330 21.351204 21.538540 21.713249
#> 8  18.56103 18.632097 18.711156 18.784892 18.853659 18.917791 18.977601
#> 9  18.74342 18.872982 19.010221 19.138219 19.257590 19.368917 19.472740
#> 10 14.77246 14.769495 14.793946 14.816751 14.838018 14.857853 14.876350
#> 11  7.34572  7.413157  7.548526  7.674781  7.792526  7.902337  8.004746
#> 12 28.81019 28.862998 28.868780 28.874172 28.879200 28.883890 28.888264
#> 13 28.00224 28.194369 28.345367 28.486198 28.617538 28.740026 28.854259
#> 14 18.28494 18.258408 18.240244 18.223302 18.207502 18.192767 18.179025
#> 15 15.45077 15.248800 15.068367 14.900083 14.743140 14.596775 14.460274
#> 16 15.92141 15.879433 15.858282 15.838555 15.820157 15.803000 15.786999
#> 17 27.14032 27.237135 27.296352 27.351582 27.403090 27.451126 27.495925
#> 18 18.91229 18.607033 18.303628 18.020652 17.756747 17.510628 17.281098
#> 19 21.44801 21.231539 21.004363 20.792483 20.594882 20.410600 20.238737
#> 20 16.71096 16.424150 16.151178 15.896586 15.659152 15.437720 15.231213
#> 21 22.33458 22.420540 22.494474 22.563429 22.627738 22.687712 22.743644
#>       0.9659    0.9008   0.8401    0.7835    0.7307    0.6814    0.6355
#> 1  21.730521 21.637176 21.55012 21.468936 21.393221 21.322609 21.308699
#> 2  20.880878 20.831960 20.78634 20.743792 20.704113 20.667108 20.659819
#> 3  25.240187 25.273781 25.30511 25.334330 25.361579 25.386991 25.391997
#> 4  21.243383 21.360461 21.46965 21.571476 21.666442 21.755007 21.772452
#> 5  20.643286 20.818485 20.98188 21.134257 21.276367 21.408899 21.435005
#> 6  16.146740 16.046243 15.95252 15.865113 15.783597 15.707575 15.692600
#> 7  21.876184 22.028138 22.16985 22.302012 22.425266 22.540213 22.562856
#> 8  19.033380 19.085399 19.13391 19.179157 19.221351 19.260702 19.268454
#> 9  19.569566 19.659866 19.74408 19.822618 19.895864 19.964172 19.977628
#> 10 14.893601 14.909689 14.92469 14.938686 14.951735 14.963906 14.966303
#> 11  8.100253  8.189323  8.27239  8.349859  8.422107  8.489485  8.502758
#> 12 28.892343 28.896147 28.89969 28.903003 28.906089 28.908966 28.909533
#> 13 28.960792 29.060146 29.15280 29.239217 29.319806 29.394963 29.409768
#> 14 18.166209 18.154257 18.14311 18.132715 18.123020 18.113979 18.112198
#> 15 14.332974 14.214252 14.10353 14.000275 13.903976 13.814168 13.796478
#> 16 15.772076 15.758159 15.74518 15.733076 15.721787 15.711259 15.709186
#> 17 27.537705 27.576669 27.61301 27.646895 27.678500 27.707975 27.713781
#> 18 17.067036 16.867402 16.68122 16.507591 16.345661 16.194645 16.164898
#> 19 20.078457 19.928980 19.78958 19.659569 19.538323 19.425249 19.402975
#> 20 15.038623 14.859013 14.69151 14.535292 14.389605 14.253737 14.226974
#> 21 22.795806 22.844453 22.88982 22.932132 22.971591 23.008390 23.015639
#>       0.5927    0.5527    0.5155    0.4807    0.4483    0.4181    0.3899
#> 1  21.308695 21.308695 21.308695 21.308695 21.308695 21.308695 21.308695
#> 2  20.659816 20.659816 20.659816 20.659816 20.659816 20.659816 20.659816
#> 3  25.391999 25.391999 25.391999 25.391999 25.391999 25.391999 25.391999
#> 4  21.772458 21.772458 21.772458 21.772458 21.772458 21.772458 21.772458
#> 5  21.435013 21.435013 21.435013 21.435013 21.435013 21.435013 21.435013
#> 6  15.692596 15.692596 15.692596 15.692596 15.692596 15.692596 15.692596
#> 7  22.562863 22.562863 22.562863 22.562863 22.562863 22.562863 22.562863
#> 8  19.268456 19.268456 19.268456 19.268456 19.268456 19.268456 19.268456
#> 9  19.977632 19.977632 19.977632 19.977632 19.977632 19.977632 19.977632
#> 10 14.966304 14.966304 14.966304 14.966304 14.966304 14.966304 14.966304
#> 11  8.502761  8.502762  8.502762  8.502762  8.502762  8.502762  8.502762
#> 12 28.909533 28.909533 28.909533 28.909533 28.909533 28.909533 28.909533
#> 13 29.409772 29.409772 29.409772 29.409772 29.409772 29.409772 29.409772
#> 14 18.112198 18.112198 18.112198 18.112198 18.112198 18.112198 18.112198
#> 15 13.796472 13.796472 13.796472 13.796472 13.796472 13.796472 13.796472
#> 16 15.709185 15.709185 15.709185 15.709185 15.709185 15.709185 15.709185
#> 17 27.713782 27.713782 27.713782 27.713782 27.713782 27.713782 27.713782
#> 18 16.164889 16.164889 16.164889 16.164889 16.164889 16.164889 16.164889
#> 19 19.402969 19.402968 19.402968 19.402968 19.402968 19.402968 19.402968
#> 20 14.226966 14.226965 14.226965 14.226965 14.226965 14.226965 14.226965
#> 21 23.015641 23.015642 23.015642 23.015642 23.015642 23.015642 23.015642
#>       0.3637    0.3391    0.3163    0.2950   0.2751   0.2566   0.2393   0.2231
#> 1  21.308695 21.308695 21.308695 21.283835 22.33912 22.33924 22.33934 22.33958
#> 2  20.659816 20.659816 20.659816 20.674377 22.39822 22.39851 22.39875 22.39986
#> 3  25.391999 25.391999 25.391999 25.349671 24.35835 24.35818 24.35804 24.35768
#> 4  21.772458 21.772458 21.772458 21.658578 20.12611 20.12588 20.12570 20.12413
#> 5  21.435013 21.435013 21.435013 21.282307 19.93511 19.93495 19.93483 19.93449
#> 6  15.692596 15.692596 15.692596 15.601572 14.80224 14.80199 14.80179 14.79909
#> 7  22.562863 22.562863 22.562863 22.607212 23.00055 23.00075 23.00092 23.00298
#> 8  19.268456 19.268456 19.268456 19.375957 18.59345 18.59337 18.59330 18.59346
#> 9  19.977632 19.977632 19.977632 20.093988 19.93597 19.93604 19.93610 19.93697
#> 10 14.966304 14.966304 14.966304 14.924274 15.60651 15.60664 15.60674 15.60636
#> 11  8.502762  8.502762  8.502762  8.614806 10.03414 10.03451 10.03480 10.03604
#> 12 28.909533 28.909533 28.909533 29.050658 30.12129 30.12153 30.12172 30.12030
#> 13 29.409772 29.409772 29.409772 29.420039 30.42805 30.42832 30.42854 30.42920
#> 14 18.112198 18.112198 18.112198 18.011175 18.55315 18.55322 18.55327 18.55223
#> 15 13.796472 13.796472 13.796472 13.862534 13.85836 13.85829 13.85822 13.85505
#> 16 15.709185 15.709185 15.709185 15.637008 15.90149 15.90151 15.90153 15.90055
#> 17 27.713782 27.713782 27.713782 27.686390 27.37704 27.37701 27.37698 27.37658
#> 18 16.164889 16.164889 16.164889 16.258575 16.16761 16.16746 16.16734 16.16878
#> 19 19.402968 19.402968 19.402968 19.312030 18.79825 18.79802 18.79783 18.80036
#> 20 14.226965 14.226965 14.226965 14.200422 12.84154 12.84114 12.84081 12.84294
#> 21 23.015642 23.015642 23.015642 23.094591 22.82344 22.82344 22.82345 22.82336
#>      0.2081   0.1941   0.1810   0.1688   0.1574   0.1468   0.1369   0.1277
#> 1  22.34856 22.35700 21.85184 21.97581 21.97579 21.97578 21.97577 21.97576
#> 2  22.45820 22.51263 22.09892 22.24031 22.24031 22.24031 22.24030 22.24030
#> 3  24.35006 24.34298 24.15691 23.32428 23.32403 23.32379 23.32356 23.32335
#> 4  20.05460 19.98989 20.38691 20.60501 20.60516 20.60529 20.60542 20.60554
#> 5  19.95785 19.97981 20.11394 20.20020 20.20027 20.20035 20.20042 20.20048
#> 6  14.63401 14.48022 14.28655 14.84800 14.84816 14.84831 14.84845 14.84858
#> 7  23.13917 23.26608 23.59645 23.49345 23.49345 23.49345 23.49345 23.49345
#> 8  18.59226 18.59096 17.87363 17.92471 17.92454 17.92438 17.92423 17.92408
#> 9  19.98004 20.02001 19.69267 20.14061 20.14061 20.14061 20.14061 20.14061
#> 10 15.58424 15.56369 14.62033 14.09739 14.09713 14.09689 14.09666 14.09645
#> 11 10.09367 10.14727 10.68177 10.78861 10.78870 10.78878 10.78886 10.78894
#> 12 29.99063 29.86954 30.01911 30.67095 30.67112 30.67128 30.67143 30.67158
#> 13 30.46953 30.50714 31.18188 31.61084 31.61108 31.61130 31.61151 31.61171
#> 14 18.49579 18.44333 18.52863 18.34741 18.34744 18.34747 18.34749 18.34752
#> 15 13.62190 13.40443 12.54983 12.42924 12.42907 12.42890 12.42875 12.42860
#> 16 15.84464 15.79262 16.65606 16.52681 16.52696 16.52710 16.52724 16.52737
#> 17 27.35836 27.34141 27.35946 27.03803 27.03795 27.03788 27.03782 27.03775
#> 18 16.24069 16.30762 17.81830 16.90910 16.90906 16.90902 16.90898 16.90895
#> 19 18.98519 19.15767 18.71284 18.67591 18.67584 18.67577 18.67571 18.67565
#> 20 12.99351 13.13393 13.49303 14.45866 14.45894 14.45921 14.45946 14.45969
#> 21 22.80708 22.79178 22.32096 21.69468 21.69439 21.69413 21.69387 21.69363
#>      0.1191   0.1111   0.1036   0.0966   0.0901   0.0840   0.0783   0.0731
#> 1  21.97574 21.97573 21.97572 21.97571 21.97570 21.97570 21.97569 21.97568
#> 2  22.24030 22.24029 22.24029 22.24029 22.24029 22.24028 22.24028 22.24028
#> 3  23.32315 23.32296 23.32278 23.32261 23.32245 23.32230 23.32216 23.32202
#> 4  20.60566 20.60577 20.60587 20.60597 20.60606 20.60614 20.60623 20.60630
#> 5  20.20054 20.20060 20.20065 20.20071 20.20075 20.20080 20.20084 20.20088
#> 6  14.84871 14.84883 14.84894 14.84905 14.84915 14.84925 14.84933 14.84942
#> 7  23.49345 23.49344 23.49344 23.49344 23.49344 23.49344 23.49344 23.49344
#> 8  17.92395 17.92382 17.92370 17.92358 17.92347 17.92337 17.92327 17.92318
#> 9  20.14061 20.14061 20.14062 20.14062 20.14062 20.14062 20.14062 20.14062
#> 10 14.09624 14.09605 14.09587 14.09570 14.09554 14.09539 14.09524 14.09511
#> 11 10.78901 10.78908 10.78914 10.78920 10.78926 10.78931 10.78936 10.78941
#> 12 30.67171 30.67184 30.67196 30.67208 30.67219 30.67229 30.67239 30.67248
#> 13 31.61190 31.61207 31.61224 31.61240 31.61255 31.61269 31.61282 31.61295
#> 14 18.34754 18.34757 18.34759 18.34761 18.34763 18.34765 18.34767 18.34768
#> 15 12.42846 12.42833 12.42821 12.42809 12.42798 12.42788 12.42778 12.42769
#> 16 16.52749 16.52761 16.52772 16.52782 16.52791 16.52801 16.52809 16.52817
#> 17 27.03769 27.03764 27.03758 27.03753 27.03749 27.03744 27.03740 27.03736
#> 18 16.90892 16.90889 16.90886 16.90883 16.90881 16.90878 16.90876 16.90874
#> 19 18.67559 18.67554 18.67549 18.67544 18.67539 18.67535 18.67531 18.67527
#> 20 14.45992 14.46013 14.46033 14.46051 14.46069 14.46086 14.46101 14.46116
#> 21 21.69341 21.69320 21.69300 21.69281 21.69263 21.69246 21.69230 21.69215
#>      0.0681   0.0636   0.0593   0.0553   0.0515   0.0481   0.0448   0.0418
#> 1  21.34201 21.34183 21.34166 21.34149 21.34133 21.34118 21.34103 21.34089
#> 2  21.79397 21.79396 21.79394 21.79393 21.79392 21.79390 21.79389 21.79388
#> 3  23.50125 23.50122 23.50119 23.50117 23.50114 23.50111 23.50109 23.50107
#> 4  20.06642 20.06616 20.06592 20.06569 20.06546 20.06525 20.06504 20.06484
#> 5  19.78175 19.78171 19.78167 19.78164 19.78160 19.78157 19.78154 19.78151
#> 6  15.60845 15.60835 15.60825 15.60816 15.60807 15.60799 15.60791 15.60783
#> 7  24.87640 24.87653 24.87666 24.87678 24.87690 24.87702 24.87713 24.87723
#> 8  17.19232 17.19227 17.19222 17.19217 17.19213 17.19209 17.19204 17.19201
#> 9  19.72205 19.72213 19.72222 19.72230 19.72238 19.72246 19.72253 19.72260
#> 10 14.19786 14.19821 14.19855 14.19888 14.19920 14.19951 14.19980 14.20008
#> 11 10.71945 10.71936 10.71928 10.71920 10.71912 10.71904 10.71897 10.71890
#> 12 30.39125 30.39114 30.39103 30.39093 30.39083 30.39073 30.39064 30.39055
#> 13 32.21188 32.21216 32.21244 32.21270 32.21295 32.21319 32.21343 32.21365
#> 14 18.09178 18.09190 18.09201 18.09211 18.09221 18.09231 18.09240 18.09249
#> 15 13.10877 13.10882 13.10887 13.10891 13.10895 13.10899 13.10903 13.10907
#> 16 16.33308 16.33293 16.33280 16.33266 16.33254 16.33241 16.33230 16.33218
#> 17 27.01817 27.01816 27.01816 27.01815 27.01815 27.01814 27.01814 27.01813
#> 18 16.56410 16.56419 16.56428 16.56437 16.56446 16.56454 16.56462 16.56470
#> 19 18.82777 18.82767 18.82758 18.82749 18.82740 18.82732 18.82723 18.82716
#> 20 14.71338 14.71337 14.71335 14.71334 14.71333 14.71332 14.71332 14.71331
#> 21 21.93792 21.93792 21.93792 21.93793 21.93793 21.93793 21.93793 21.93794
#>      0.0390   0.0364   0.0339   0.0316   0.0295   0.0275   0.0257   0.0239
#> 1  21.34076 21.34063 21.34050 21.34038 21.34027 21.34016 21.34005 21.33995
#> 2  21.79387 21.79386 21.79385 21.79384 21.79383 21.79382 21.79381 21.79380
#> 3  23.50104 23.50102 23.50100 23.50098 23.50096 23.50094 23.50093 23.50091
#> 4  20.06465 20.06447 20.06429 20.06412 20.06396 20.06381 20.06366 20.06351
#> 5  19.78148 19.78146 19.78143 19.78141 19.78138 19.78136 19.78134 19.78132
#> 6  15.60775 15.60768 15.60761 15.60754 15.60747 15.60741 15.60735 15.60729
#> 7  24.87733 24.87743 24.87752 24.87761 24.87769 24.87778 24.87785 24.87793
#> 8  17.19197 17.19193 17.19190 17.19187 17.19183 17.19180 17.19178 17.19175
#> 9  19.72267 19.72273 19.72279 19.72285 19.72291 19.72296 19.72302 19.72307
#> 10 14.20035 14.20061 14.20086 14.20110 14.20133 14.20155 14.20176 14.20196
#> 11 10.71883 10.71877 10.71871 10.71865 10.71859 10.71854 10.71848 10.71843
#> 12 30.39047 30.39039 30.39031 30.39024 30.39016 30.39009 30.39003 30.38996
#> 13 32.21386 32.21407 32.21427 32.21446 32.21464 32.21481 32.21498 32.21514
#> 14 18.09258 18.09266 18.09274 18.09282 18.09289 18.09296 18.09303 18.09309
#> 15 13.10910 13.10914 13.10917 13.10920 13.10923 13.10926 13.10929 13.10931
#> 16 16.33207 16.33197 16.33187 16.33177 16.33168 16.33159 16.33151 16.33142
#> 17 27.01813 27.01813 27.01812 27.01812 27.01812 27.01812 27.01811 27.01811
#> 18 16.56477 16.56484 16.56491 16.56497 16.56503 16.56510 16.56515 16.56521
#> 19 18.82708 18.82701 18.82694 18.82687 18.82681 18.82675 18.82669 18.82663
#> 20 14.71330 14.71329 14.71328 14.71327 14.71326 14.71326 14.71325 14.71324
#> 21 21.93794 21.93794 21.93794 21.93794 21.93795 21.93795 21.93795 21.93795
#>      0.0223   0.0208   0.0194   0.0181   0.0169   0.0157   0.0147   0.0137
#> 1  21.33985 21.33976 21.33967 21.33958 21.33950 21.33942 21.33934 21.33927
#> 2  21.79379 21.79378 21.79378 21.79377 21.79376 21.79376 21.79375 21.79374
#> 3  23.50089 23.50088 23.50086 23.50085 23.50083 23.50082 23.50081 23.50080
#> 4  20.06337 20.06324 20.06311 20.06299 20.06288 20.06276 20.06265 20.06255
#> 5  19.78130 19.78128 19.78126 19.78124 19.78123 19.78121 19.78119 19.78118
#> 6  15.60723 15.60718 15.60713 15.60708 15.60703 15.60698 15.60694 15.60690
#> 7  24.87800 24.87807 24.87813 24.87820 24.87826 24.87832 24.87837 24.87842
#> 8  17.19172 17.19170 17.19167 17.19165 17.19163 17.19161 17.19159 17.19157
#> 9  19.72312 19.72317 19.72321 19.72325 19.72330 19.72334 19.72337 19.72341
#> 10 14.20216 14.20235 14.20253 14.20270 14.20287 14.20303 14.20318 14.20333
#> 11 10.71838 10.71834 10.71829 10.71825 10.71821 10.71817 10.71813 10.71809
#> 12 30.38990 30.38984 30.38979 30.38973 30.38968 30.38963 30.38958 30.38954
#> 13 32.21530 32.21545 32.21559 32.21573 32.21586 32.21598 32.21610 32.21622
#> 14 18.09316 18.09322 18.09328 18.09333 18.09338 18.09344 18.09349 18.09353
#> 15 13.10934 13.10936 13.10939 13.10941 13.10943 13.10945 13.10947 13.10949
#> 16 16.33134 16.33127 16.33120 16.33113 16.33106 16.33099 16.33093 16.33087
#> 17 27.01811 27.01811 27.01810 27.01810 27.01810 27.01810 27.01809 27.01809
#> 18 16.56526 16.56531 16.56536 16.56541 16.56546 16.56550 16.56555 16.56559
#> 19 18.82658 18.82652 18.82647 18.82642 18.82638 18.82633 18.82629 18.82625
#> 20 14.71324 14.71323 14.71322 14.71322 14.71321 14.71320 14.71320 14.71319
#> 21 21.93795 21.93795 21.93796 21.93796 21.93796 21.93796 21.93796 21.93796
#>      0.0128   0.0119   0.0111   0.0104   0.0097   0.0090   0.0084   0.0078
#> 1  21.33920 21.33913 21.33907 21.33900 21.33894 21.33889 21.33883 21.33878
#> 2  21.79374 21.79373 21.79373 21.79372 21.79372 21.79371 21.79371 21.79370
#> 3  23.50079 23.50077 23.50076 23.50075 23.50074 23.50073 23.50073 23.50072
#> 4  20.06245 20.06236 20.06226 20.06218 20.06209 20.06201 20.06193 20.06186
#> 5  19.78116 19.78115 19.78114 19.78113 19.78111 19.78110 19.78109 19.78108
#> 6  15.60686 15.60682 15.60678 15.60674 15.60671 15.60667 15.60664 15.60661
#> 7  24.87848 24.87852 24.87857 24.87862 24.87866 24.87870 24.87874 24.87878
#> 8  17.19155 17.19153 17.19152 17.19150 17.19148 17.19147 17.19146 17.19144
#> 9  19.72345 19.72348 19.72352 19.72355 19.72358 19.72361 19.72364 19.72366
#> 10 14.20347 14.20361 14.20374 14.20387 14.20399 14.20410 14.20421 14.20432
#> 11 10.71806 10.71802 10.71799 10.71796 10.71793 10.71790 10.71787 10.71784
#> 12 30.38949 30.38945 30.38941 30.38937 30.38933 30.38930 30.38926 30.38923
#> 13 32.21633 32.21644 32.21654 32.21664 32.21673 32.21682 32.21691 32.21700
#> 14 18.09358 18.09362 18.09366 18.09370 18.09374 18.09378 18.09382 18.09385
#> 15 13.10951 13.10952 13.10954 13.10956 13.10957 13.10959 13.10960 13.10961
#> 16 16.33082 16.33076 16.33071 16.33066 16.33061 16.33056 16.33052 16.33048
#> 17 27.01809 27.01809 27.01809 27.01809 27.01809 27.01808 27.01808 27.01808
#> 18 16.56563 16.56566 16.56570 16.56573 16.56577 16.56580 16.56583 16.56586
#> 19 18.82621 18.82617 18.82613 18.82610 18.82606 18.82603 18.82600 18.82597
#> 20 14.71319 14.71318 14.71318 14.71317 14.71317 14.71316 14.71316 14.71316
#> 21 21.93796 21.93797 21.93797 21.93797 21.93797 21.93797 21.93797 21.93797
#>      0.0073   0.0068   0.0064   0.0059   0.0055   0.0052   0.0048
#> 1  21.33873 21.33868 21.33863 21.33859 21.33854 21.33850 21.33846
#> 2  21.79370 21.79370 21.79369 21.79369 21.79369 21.79368 21.79368
#> 3  23.50071 23.50070 23.50069 23.50069 23.50068 23.50067 23.50067
#> 4  20.06179 20.06172 20.06165 20.06159 20.06153 20.06147 20.06141
#> 5  19.78107 19.78106 19.78105 19.78104 19.78103 19.78103 19.78102
#> 6  15.60658 15.60655 15.60652 15.60650 15.60647 15.60644 15.60642
#> 7  24.87881 24.87885 24.87888 24.87891 24.87894 24.87897 24.87900
#> 8  17.19143 17.19142 17.19140 17.19139 17.19138 17.19137 17.19136
#> 9  19.72369 19.72371 19.72374 19.72376 19.72378 19.72380 19.72382
#> 10 14.20442 14.20452 14.20462 14.20471 14.20480 14.20488 14.20496
#> 11 10.71782 10.71779 10.71777 10.71775 10.71772 10.71770 10.71768
#> 12 30.38920 30.38917 30.38914 30.38911 30.38908 30.38905 30.38903
#> 13 32.21708 32.21715 32.21723 32.21730 32.21737 32.21743 32.21749
#> 14 18.09388 18.09392 18.09395 18.09398 18.09400 18.09403 18.09406
#> 15 13.10963 13.10964 13.10965 13.10966 13.10967 13.10968 13.10969
#> 16 16.33043 16.33039 16.33036 16.33032 16.33028 16.33025 16.33022
#> 17 27.01808 27.01808 27.01808 27.01808 27.01808 27.01808 27.01807
#> 18 16.56589 16.56592 16.56595 16.56597 16.56600 16.56602 16.56604
#> 19 18.82594 18.82591 18.82589 18.82586 18.82584 18.82581 18.82579
#> 20 14.71315 14.71315 14.71314 14.71314 14.71314 14.71313 14.71313
#> 21 21.93797 21.93797 21.93798 21.93798 21.93798 21.93798 21.93798
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 22.8 21.4 18.1 14.3 24.4 19.2 17.8 16.4 10.4 30.4 33.9 15.2 13.3
#> [16] 19.2 27.3 15.8 19.7 15.0 21.4
#> 
#> $X
#>               am        carb       cyl        disp        drat       gear
#>  [1,]  1.0488088  0.53338075 -0.117444 -0.55650168  0.45145215  0.3483665
#>  [2,]  1.0488088  0.53338075 -0.117444 -0.55650168  0.45145215  0.3483665
#>  [3,]  1.0488088 -1.14676861 -1.350607 -1.01462444  0.35510565  0.3483665
#>  [4,] -0.9534626 -1.14676861 -0.117444  0.30688351 -1.12863037 -1.1147728
#>  [5,] -0.9534626 -1.14676861 -0.117444  0.01615176 -1.74524794 -1.1147728
#>  [6,] -0.9534626  0.53338075  1.115718  1.20550892 -0.87812948 -1.1147728
#>  [7,] -0.9534626 -0.58671882 -1.350607 -0.67367539  0.04679687  0.3483665
#>  [8,] -0.9534626  0.53338075 -0.117444 -0.48954528  0.48999075  0.3483665
#>  [9,] -0.9534626  0.53338075 -0.117444 -0.48954528  0.48999075  0.3483665
#> [10,] -0.9534626 -0.02666904  1.115718  0.46370246 -1.14789967 -1.1147728
#> [11,] -0.9534626  0.53338075  1.115718  2.08651422 -1.28278476 -1.1147728
#> [12,]  1.0488088 -0.58671882 -1.350607 -1.29918915  2.43618995  0.3483665
#> [13,]  1.0488088 -1.14676861 -1.350607 -1.33971539  1.06806972  0.3483665
#> [14,] -0.9534626 -0.58671882  1.115718  0.71214595 -0.99374528 -1.1147728
#> [15,] -0.9534626  0.53338075  1.115718  1.11740839  0.12387407 -1.1147728
#> [16,] -0.9534626 -0.58671882  1.115718  1.55791104 -1.12863037 -1.1147728
#> [17,]  1.0488088 -1.14676861 -1.350607 -1.27011597  0.79829953  0.3483665
#> [18,]  1.0488088  0.53338075  1.115718  1.12621844  1.06806972  1.8115058
#> [19,]  1.0488088  1.65348032 -0.117444 -0.68865248 -0.08808822  1.8115058
#> [20,]  1.0488088  2.77357988  1.115718  0.68571579 -0.24224262  1.8115058
#> [21,]  1.0488088 -0.58671882 -1.350607 -0.90009375  0.85610743  0.3483665
#>                hp       qsec         vs           wt
#>  [1,] -0.51292320 -0.6879102 -0.9534626 -0.651756520
#>  [2,] -0.51292320 -0.3594561 -0.9534626 -0.345575613
#>  [3,] -0.74124438  0.5731188  1.0488088 -1.011969353
#>  [4,] -0.51292320  1.0599347  1.0488088  0.062665598
#>  [5,] -0.58007649  1.5174243  1.0488088  0.356839411
#>  [6,]  1.30021555 -1.0515558 -0.9534626  0.488917449
#>  [7,] -1.15759476  1.3883888  1.0488088  0.032647861
#>  [8,] -0.33832466  0.3912961  1.0488088  0.332825222
#>  [9,] -0.33832466  0.7432111  1.0488088  0.332825222
#> [10,]  0.42722282 -0.1365766 -0.9534626  1.089272170
#> [11,]  0.89729583  0.1097640 -0.9534626  2.715032754
#> [12,] -1.29190133  0.5203316  1.0488088 -1.858469509
#> [13,] -1.11730278  1.3297363  1.0488088 -1.594313432
#> [14,]  0.02430309 -0.1952291 -0.9534626  0.326821675
#> [15,]  1.30021555 -1.3037616 -0.9534626  0.813108999
#> [16,]  0.36006953 -0.3418604 -0.9534626  0.819112546
#> [17,] -1.10387213  0.7432111  1.0488088 -1.474242488
#> [18,]  1.55539804 -1.8374995 -0.9534626  0.008633673
#> [19,]  0.36006953 -1.2509743 -0.9534626 -0.471650104
#> [20,]  2.50897471 -1.7788470 -0.9534626  0.488917449
#> [21,] -0.52635386  0.5672536  1.0488088 -0.459643010
#> attr(,"center")
#>  [1]   0.4761905   3.0476190   6.1904762 223.1666667   3.6657143   3.7619048
#>  [7] 148.1904762  17.6328571   0.4761905   3.1628095
#> attr(,"scale")
#>  [1]   0.4994328   1.7855555   1.6218463 113.5066951   0.5189602   0.6834619
#>  [7]  74.4565191   1.7049568   0.4994328   0.8328410
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 47
#> 
#> $lambda.min
#> [1] 0.1940732
#> 
#> $null.dev
#> [1] 30.74236
#> 
#> $Bias
#> [1] 3.510494
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 16.70516