Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 41.50146 38.24270 33.64911 29.39870 25.94027 23.15462 20.93912 19.20554
#>   [9] 17.87809 16.89169 16.19061 15.72711 15.46038 15.35561 15.38312 15.51769
#>  [17] 15.35515 15.09253 14.99086 14.88288 14.71385 14.35174 13.98189 14.05729
#>  [25] 14.30073 14.51221 14.81116 14.67729 14.22242 13.54208 12.94358 12.15472
#>  [33] 11.65285 10.82037 13.79973 13.63761 13.60122 13.46485 13.40096 13.50219
#>  [41] 14.14546 12.41037 12.54795 12.31623 12.22067 11.97400 11.13211 13.58520
#>  [49] 19.07999 19.14811 19.14942 19.12269 18.75232 18.69637 18.48082 18.67041
#>  [57] 18.69579 19.01176 20.71777 20.48475 20.47104 20.15283 20.15321 21.21004
#>  [65] 21.09827 22.03743 21.97758 21.97765 21.98341 21.43432 22.00253 22.00249
#>  [73] 22.00238 22.18727 22.18047 22.25553 22.31378 22.31481 22.31581 22.31679
#>  [81] 22.35894 22.35990 22.47476 22.47465 22.47383 22.47369 22.47092 22.47091
#>  [89] 22.47091 22.47092 22.47093 22.47094 22.47096 22.47098 22.47101 22.47104
#>  [97] 22.47108 22.47112 22.47117 22.47121
#> 
#> $cvse
#>   [1] 13.502098 13.401099 12.415319 11.219860 10.155895  9.204342  8.350766
#>   [8]  7.584631  6.898655  6.288251  5.751016  5.286216  4.894186  4.575590
#>  [15]  4.330537  4.157679  3.955465  3.758359  3.735269  3.735721  3.736468
#>  [22]  3.713885  3.692262  3.682012  3.700448  3.769439  3.909704  3.940969
#>  [29]  3.922355  3.798862  3.619369  3.370488  3.208644  3.072016  4.848759
#>  [36]  4.843143  4.840756  4.767583  4.746398  4.765350  4.714418  3.683729
#>  [43]  3.673036  3.693049  3.689706  3.694262  3.222217  3.968584  6.061887
#>  [50]  6.066054  6.066618  6.064418  6.090386  6.090519  6.099452  6.075430
#>  [57]  6.068095  6.149888  6.817018  6.785474  6.786434  6.610053  6.610418
#>  [64]  7.117667  7.123904  7.804711  7.808761  7.808380  7.807324  7.715148
#>  [71]  7.665829  7.665478  7.665150  7.674169  7.674932  7.664725  7.658454
#>  [78]  7.658328  7.658217  7.658121  7.652275  7.652201  7.725789  7.725334
#>  [85]  7.724989  7.724592  7.724633  7.724266  7.723920  7.723595  7.723289
#>  [92]  7.723002  7.722731  7.722477  7.722239  7.722015  7.721804  7.721607
#>  [99]  7.721422  7.721248
#> 
#> $fold
#>  [1]  3  1 10  5  6  9  9  2  4  4 10  2  6  3  8  7  5  7  1  8  1
#> 
#> $lambda
#>   [1] 5.339540978 4.979673787 4.644060440 4.331066310 4.039166936 3.766940603
#>   [7] 3.513061414 3.276292832 3.055481661 2.849552423 2.657502127 2.478395378
#>  [13] 2.311359825 2.155581908 2.010302902 1.874815215 1.748458944 1.630618663
#>  [19] 1.520720423 1.418228956 1.322645071 1.233503220 1.150369231 1.072838195
#>  [25] 1.000532491 0.933099950 0.870212137 0.811562751 0.756866138 0.705855893
#>  [31] 0.658283569 0.613917459 0.572541477 0.533954098 0.497967379 0.464406044
#>  [37] 0.433106631 0.403916694 0.376694060 0.351306141 0.327629283 0.305548166
#>  [43] 0.284955242 0.265750213 0.247839538 0.231135982 0.215558191 0.201030290
#>  [49] 0.187481521 0.174845894 0.163061866 0.152072042 0.141822896 0.132264507
#>  [55] 0.123350322 0.115036923 0.107283820 0.100053249 0.093309995 0.087021214
#>  [61] 0.081156275 0.075686614 0.070585589 0.065828357 0.061391746 0.057254148
#>  [67] 0.053395410 0.049796738 0.046440604 0.043310663 0.040391669 0.037669406
#>  [73] 0.035130614 0.032762928 0.030554817 0.028495524 0.026575021 0.024783954
#>  [79] 0.023113598 0.021555819 0.020103029 0.018748152 0.017484589 0.016306187
#>  [85] 0.015207204 0.014182290 0.013226451 0.012335032 0.011503692 0.010728382
#>  [91] 0.010005325 0.009331000 0.008702121 0.008115628 0.007568661 0.007058559
#>  [97] 0.006582836 0.006139175 0.005725415 0.005339541
#> 
#> $fit
#> $beta
#>               5.3395       4.9797       4.6441      4.3311      4.0392
#> (Intercept) 19.58095 20.536235241 21.490294008 22.38005241 23.20984407
#> am           0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> carb         0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> cyl          0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> disp         0.00000 -0.003952583 -0.007900101 -0.01158157 -0.01501492
#> drat         0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> gear         0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> hp           0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> qsec         0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> vs           0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> wt           0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#>                  3.7669      3.5131      3.2763      3.0555      2.8496
#> (Intercept) 23.98371055 24.70542102 25.37849062 26.00619758 26.59159919
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.01821687 -0.02120303 -0.02398792 -0.02658513 -0.02900729
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                 2.6575      2.4784      2.3114      2.1556      2.0103
#> (Intercept) 27.1375467 27.64669916 28.12153646 28.56437131 28.97736058
#> am           0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.0312662 -0.03337287 -0.03533756 -0.03716984 -0.03887862
#> drat         0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  1.8748      1.7485      1.6306      1.5207      1.4182
#> (Intercept) 29.36251575 29.65505930 29.65505930 29.65505930 29.77894368
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000 -0.08141798
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.04047224 -0.04168267 -0.04168267 -0.04168267 -0.04126484
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  1.3226      1.2335      1.1504      1.0728     1.0005
#> (Intercept) 30.01869115 30.24200420 30.45049044 30.64492574 30.8262567
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> carb        -0.23898228 -0.38574573 -0.52276485 -0.65054956 -0.7697220
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> disp        -0.04045623 -0.03970305 -0.03899987 -0.03834408 -0.0377325
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#>                  0.9331     0.8702      0.8116      0.7569      0.7059
#> (Intercept) 30.99536665 31.1530791 31.22692241 30.63638784 30.10274422
#> am           0.00000000  0.0000000  0.11545677  1.26901498  2.27366800
#> carb        -0.88086265 -0.9845128 -1.09742829 -1.35261062 -1.56214453
#> cyl          0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> disp        -0.03716213 -0.0366302 -0.03578186 -0.03178603 -0.02837121
#> drat         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#>                  0.6583      0.6139      0.5725      0.5340      0.4980
#> (Intercept) 29.69431032 29.31322426 28.98172624 28.98148275 28.98134423
#> am           2.83735770  3.36333459  3.82092421  3.82141786  3.82169869
#> carb        -1.60806216 -1.65092232 -1.68823572 -1.68835098 -1.68841655
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.02682292 -0.02537814 -0.02412108 -0.02411934 -0.02411835
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.4644      0.4331      0.4039      0.3767     0.3513
#> (Intercept) 28.98126543 28.98122060 28.98119509 28.98118059 15.5799002
#> am           3.82185845  3.82194933  3.82200104  3.82203045  2.4439810
#> carb        -1.68845385 -1.68847507 -1.68848714 -1.68849401 -2.4477709
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> disp        -0.02411778 -0.02411746 -0.02411728 -0.02411718 -0.0109849
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  3.6586534
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#>                  0.3276      0.3055      0.2850      0.2658      0.2478
#> (Intercept) 15.57728641 15.57508089 15.57321988 15.57164957 15.57032454
#> am           2.44394745  2.44391918  2.44389532  2.44387520  2.44385821
#> carb        -2.44799771 -2.44818907 -2.44835054 -2.44848678 -2.44860174
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.01098111 -0.01097791 -0.01097521 -0.01097293 -0.01097101
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         3.65932459  3.65989094  3.66036884  3.66077208  3.66111234
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.2311      0.2156       0.2010       0.1875       0.1748
#> (Intercept) 15.56920648 15.56826307  9.833777292  9.832802230 -7.879896412
#> am           2.44384388  2.44383179  2.119599815  2.119668150  1.259735575
#> carb        -2.44869875 -2.44878060 -2.439125899 -2.439350639 -3.212307618
#> cyl          0.00000000  0.00000000  0.000000000  0.000000000  1.206729899
#> disp        -0.01096939 -0.01096802 -0.007722848 -0.007720109 -0.008986312
#> drat         0.00000000  0.00000000  1.696112364  1.695697530  3.471221412
#> gear         3.66139944  3.66164170  3.395983846  3.396664722  5.294252226
#> hp           0.00000000  0.00000000  0.000000000  0.000000000  0.000000000
#> qsec         0.00000000  0.00000000  0.000000000  0.000000000  0.000000000
#> vs           0.00000000  0.00000000  0.000000000  0.000000000  0.000000000
#> wt           0.00000000  0.00000000  0.000000000  0.000000000  0.000000000
#>                   0.1631      0.1521       0.1418       0.1323       0.1234
#> (Intercept) -7.877011856 -7.87385204 -7.870500268 -7.867027468 -7.863493519
#> am           1.259980007  1.26022235  1.260460271  1.260691882  1.260915700
#> carb        -3.212342553 -3.21233864 -3.212302716 -3.212240761 -3.212158012
#> cyl          1.206439725  1.20615647  1.205882097  1.205618108  1.205365639
#> disp        -0.008983034 -0.00898041 -0.008978356 -0.008976793 -0.008975655
#> drat         3.470480098  3.46978488  3.469135473  3.468531124  3.467970713
#> gear         5.294473686  5.29457178  5.294566956  5.294477228  5.294318421
#> hp           0.000000000  0.00000000  0.000000000  0.000000000  0.000000000
#> qsec         0.000000000  0.00000000  0.000000000  0.000000000  0.000000000
#> vs           0.000000000  0.00000000  0.000000000  0.000000000  0.000000000
#> wt           0.000000000  0.00000000  0.000000000  0.000000000  0.000000000
#>                  0.1150       0.1073       0.1001       0.0933       0.0870
#> (Intercept) -7.85994853 -7.856434005 -7.852983834 -14.07426597 -14.08038184
#> am           1.26113059  1.261335707  1.261530473   1.84999261   1.84996747
#> carb        -3.21205902 -3.211947735 -3.211827535  -3.04564476  -3.04590396
#> cyl          1.20512550  1.204898210  1.204684070   1.45726144   1.45754708
#> disp        -0.00897488 -0.008974413 -0.008974205  -0.01048362  -0.01048176
#> drat         3.46745282  3.466975812  3.466537867   3.54416370   3.54451451
#> gear         5.29410437  5.293847098  5.293557031   4.99308218   4.99379596
#> hp           0.00000000  0.000000000  0.000000000   0.00000000   0.00000000
#> qsec         0.00000000  0.000000000  0.000000000   0.28364741   0.28369612
#> vs           0.00000000  0.000000000  0.000000000   0.00000000   0.00000000
#> wt           0.00000000  0.000000000  0.000000000   0.00000000   0.00000000
#>                   0.0812        0.0757        0.0706        0.0658
#> (Intercept) -14.08586208 -16.801657496 -16.798209286 -16.795144147
#> am            1.84993009   1.954933385   1.955085019   1.955219871
#> carb         -3.04614903  -2.638725018  -2.638604954  -2.638502169
#> cyl           1.45780057   1.431065970   1.430973535   1.430899138
#> disp         -0.01047993  -0.002467113  -0.002471699  -0.002476141
#> drat          3.54482619   3.399081067   3.398984533   3.398913511
#> gear          4.99447072   4.731849968   4.731267904   4.730738478
#> hp            0.00000000   0.000000000   0.000000000   0.000000000
#> qsec          0.28373454   0.583760028   0.583706071   0.583653992
#> vs            0.00000000   0.000000000   0.000000000   0.000000000
#> wt            0.00000000  -1.239929356  -1.239575427  -1.239227075
#>                    0.0614        0.0573        0.0534       0.0498
#> (Intercept) -16.792436835 -16.790063176 -16.788000059 -16.78622542
#> am            1.955339072   1.955443703   1.955534791   1.95561332
#> carb         -2.638415374  -2.638343341  -2.638284910  -2.63823898
#> cyl           1.430841324   1.430798715   1.430770011   1.43075398
#> disp         -0.002480436  -0.002484581  -0.002488576  -0.00249242
#> drat          3.398865808   3.398839358   3.398832209   3.39884253
#> gear          4.730258591   4.729825256   4.729435601   4.72908687
#> hp            0.000000000   0.000000000   0.000000000   0.00000000
#> qsec          0.583603816   0.583555559   0.583509226   0.58346481
#> vs            0.000000000   0.000000000   0.000000000   0.00000000
#> wt           -1.238885094  -1.238550178  -1.238222924  -1.23790385
#>                    0.0464        0.0433        0.0404        0.0377
#> (Intercept) -16.784718212 -16.783458419 -16.782426998 -16.781605879
#> am            1.955680212   1.955736358   1.955782595   1.955819713
#> carb         -2.638204508  -2.638180514  -2.638166068  -2.638160298
#> cyl           1.430749465   1.430755372   1.430770677   1.430794416
#> disp         -0.002496113  -0.002499656  -0.002503049  -0.002506294
#> drat          3.398868587   3.398908773   3.398961568   3.399025555
#> gear          4.728776413   4.728501701   4.728260310   4.728049926
#> hp            0.000000000   0.000000000   0.000000000   0.000000000
#> qsec          0.583422312   0.583381704   0.583342966   0.583306069
#> vs            0.000000000   0.000000000   0.000000000   0.000000000
#> wt           -1.237593371  -1.237291857  -1.236999590  -1.236716793
#>                    0.0351        0.0328        0.0306        0.0285
#> (Intercept) -1.677828e+01 -16.140924001 -16.136254187 -16.131900104
#> am           1.955440e+00   1.966153372   1.966161584   1.966152859
#> carb        -2.637992e+00  -2.414398093  -2.414301860  -2.414217694
#> cyl          1.430913e+00   1.485400040   1.485090884   1.484806552
#> disp        -2.512288e-03   0.000000000   0.000000000   0.000000000
#> drat         3.399250e+00   3.627186674   3.626921067   3.626685423
#> gear         4.727627e+00   4.377997267   4.377678918   4.377395534
#> hp          -9.209749e-06  -0.008877712  -0.008875815  -0.008874239
#> qsec         5.831462e-01   0.586820246   0.586747659   0.586674727
#> vs           0.000000e+00   0.000000000   0.000000000   0.000000000
#> wt          -1.236144e+00  -1.399027179  -1.399003736  -1.398972988
#>                    0.0266        0.0248        0.0231        0.0216
#> (Intercept) -16.127846563 -16.124078529 -16.120581198 -16.117340047
#> am            1.966130185   1.966096198   1.966053213   1.966003256
#> carb         -2.414144047  -2.414079542  -2.414022961  -2.413973226
#> cyl           1.484545484   1.484306183   1.484087213   1.483887199
#> disp          0.000000000   0.000000000   0.000000000   0.000000000
#> drat          3.626477015   3.626293315   3.626131984   3.625990861
#> gear          4.377143192   4.376918368   4.376717904   4.376538973
#> hp           -0.008872953  -0.008871928  -0.008871134  -0.008870548
#> qsec          0.586602181   0.586530633   0.586460598   0.586392496
#> vs            0.000000000   0.000000000   0.000000000   0.000000000
#> wt           -1.398936476  -1.398895547  -1.398851378  -1.398804984
#>                    0.0201        0.0187      0.0175        0.0163        0.0152
#> (Intercept) -16.114340889 -16.111569913 -16.1090137 -16.106659339 -16.104501387
#> am            1.965948088   1.965889237   1.9658280   1.965765543   1.965311700
#> carb         -2.413929389  -2.413890617  -2.4138562  -2.413825434  -2.413541624
#> cyl           1.483704831   1.483538861   1.4833881   1.483251437   1.483111925
#> disp          0.000000000   0.000000000   0.0000000   0.000000000   0.000000000
#> drat          3.625867953   3.625761425   3.6256696   3.625590909   3.625456196
#> gear          4.376379052   4.376235887   4.3761075   4.375992035   4.375771068
#> hp           -0.008870144  -0.008869902  -0.0088698  -0.008869821  -0.008871030
#> qsec          0.586326669   0.586263386   0.5862029   0.586145222   0.585992852
#> vs            0.000000000   0.000000000   0.0000000   0.000000000   0.002207726
#> wt           -1.398757242  -1.398708899  -1.3986606  -1.398612845  -1.398231608
#>                    0.0142        0.0132        0.0123       0.0115
#> (Intercept) -16.439717649 -16.439337698 -16.438805237 -16.43813644
#> am            2.097027600   2.097863806   2.098627846   2.09932354
#> carb         -2.361268911  -2.361085816  -2.360925354  -2.36078604
#> cyl           1.553271862   1.553243017   1.553197856   1.55313809
#> disp          0.000000000   0.000000000   0.000000000   0.00000000
#> drat          3.541295350   3.540480452   3.539713447   3.53899277
#> gear          4.316631133   4.316271682   4.315949826   4.31566331
#> hp           -0.008815365  -0.008805682  -0.008796404  -0.00878753
#> qsec          0.596028333   0.596172305   0.596301036   0.59641545
#> vs            0.340339803   0.341174541   0.341925149   0.34259667
#> wt           -1.434196836  -1.434557867  -1.434887546  -1.43518753
#>                    0.0107        0.0100        0.0093        0.0087
#> (Intercept) -16.437346503 -16.436449710 -16.435459421 -16.434388121
#> am            2.099954627   2.100524720   2.101037347   2.101495919
#> carb         -2.360666424  -2.360565144  -2.360480874  -2.360412351
#> cyl           1.553065323   1.552981058   1.552886699   1.552783558
#> disp          0.000000000   0.000000000   0.000000000   0.000000000
#> drat          3.538316839   3.537684033   3.537092733   3.536541310
#> gear          4.315409944   4.315187621   4.314994306   4.314828037
#> hp           -0.008779057  -0.008770978  -0.008763289  -0.008755983
#> qsec          0.596516424   0.596604832   0.596681500   0.596747227
#> vs            0.343193990   0.343721787   0.344184578   0.344586699
#> wt           -1.435459435  -1.435704841  -1.435925281  -1.436122240
#>                    0.0081        0.0076        0.0071        0.0066
#> (Intercept) -16.433247449 -16.432048230 -16.430800510 -16.429513583
#> am            2.101903737   2.102263988   2.102579742   2.102853954
#> carb         -2.360358371  -2.360317783  -2.360289493  -2.360272463
#> cyl           1.552672856   1.552555728   1.552433227   1.552306327
#> disp          0.000000000   0.000000000   0.000000000   0.000000000
#> drat          3.536028136   3.535551592   3.535110069   3.534701976
#> gear          4.314686934   4.314569193   4.314473089   4.314396971
#> hp           -0.008749051  -0.008742485  -0.008736276  -0.008730414
#> qsec          0.596802779   0.596848892   0.596886268   0.596915580
#> vs            0.344932308   0.345225386   0.345469742   0.345669010
#> wt           -1.436297156  -1.436451418  -1.436586364  -1.436703283
#>                   0.0061        0.0057        0.0053
#> (Intercept) -16.42819603 -16.426855753 -16.425499999
#> am            2.10308946   2.103288979   2.103455115
#> carb         -2.36026571  -2.360268290  -2.360279333
#> cyl           1.55217593   1.552042862   1.551907885
#> disp          0.00000000   0.000000000   0.000000000
#> drat          3.53432574   3.533979828   3.533662714
#> gear          4.31433927   4.314298487   4.314273202
#> hp           -0.00872489  -0.008719693  -0.008714812
#> qsec          0.59693747   0.596952539   0.596961373
#> vs            0.34582666   0.345945981   0.346030120
#> wt           -1.43680341  -1.436887934  -1.436957988
#> 
#> $iter
#>   [1]   0   4   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   7   7   6   6   6   6   6   6  22  29  22  18  18  13   1   1   1
#>  [37]   1   1   1  49   1   1   1   1   1   1   1  42   1  64   1   1   1   1
#>  [55]   1   1   1   1  61   1   1 102   1   1   1   1   1   1   1   1   1   1
#>  [73]   2 109   1   1   1   1   1   1   1   1   1   1   2  49   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.339540978 4.979673787 4.644060440 4.331066310 4.039166936 3.766940603
#>   [7] 3.513061414 3.276292832 3.055481661 2.849552423 2.657502127 2.478395378
#>  [13] 2.311359825 2.155581908 2.010302902 1.874815215 1.748458944 1.630618663
#>  [19] 1.520720423 1.418228956 1.322645071 1.233503220 1.150369231 1.072838195
#>  [25] 1.000532491 0.933099950 0.870212137 0.811562751 0.756866138 0.705855893
#>  [31] 0.658283569 0.613917459 0.572541477 0.533954098 0.497967379 0.464406044
#>  [37] 0.433106631 0.403916694 0.376694060 0.351306141 0.327629283 0.305548166
#>  [43] 0.284955242 0.265750213 0.247839538 0.231135982 0.215558191 0.201030290
#>  [49] 0.187481521 0.174845894 0.163061866 0.152072042 0.141822896 0.132264507
#>  [55] 0.123350322 0.115036923 0.107283820 0.100053249 0.093309995 0.087021214
#>  [61] 0.081156275 0.075686614 0.070585589 0.065828357 0.061391746 0.057254148
#>  [67] 0.053395410 0.049796738 0.046440604 0.043310663 0.040391669 0.037669406
#>  [73] 0.035130614 0.032762928 0.030554817 0.028495524 0.026575021 0.024783954
#>  [79] 0.023113598 0.021555819 0.020103029 0.018748152 0.017484589 0.016306187
#>  [85] 0.015207204 0.014182290 0.013226451 0.012335032 0.011503692 0.010728382
#>  [91] 0.010005325 0.009331000 0.008702121 0.008115628 0.007568661 0.007058559
#>  [97] 0.006582836 0.006139175 0.005725415 0.005339541
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 27
#> 
#> $loss
#>   [1] 786.69238 679.49028 583.07651 502.75294 436.18575 381.36115 336.54259
#>   [8] 300.23368 271.14595 248.17089 230.35569 216.88219 207.04867 200.25397
#>  [15] 195.98379 193.79882 193.29842 193.29842 193.29842 186.59245 174.77503
#>  [22] 165.14398 157.35057 151.12510 146.22610 142.44603 139.60676 135.49435
#>  [29] 117.32542 107.63551 104.45241 102.75994 102.29083 102.29083 102.29083
#>  [36] 102.29082 102.29082 102.29082 102.29082  85.93592  85.93591  85.93590
#>  [43]  85.93590  85.93589  85.93589  85.93589  85.93589  82.48984  82.48983
#>  [50]  74.20716  74.20714  74.20713  74.20712  74.20712  74.20711  74.20710
#>  [57]  74.20710  74.20709  73.19889  73.19887  73.19886  71.14317  71.14315
#>  [64]  71.14314  71.14313  71.14312  71.14311  71.14310  71.14309  71.14309
#>  [71]  71.14308  71.14307  71.14224  70.76492  70.76491  70.76490  70.76490
#>  [78]  70.76489  70.76489  70.76489  70.76488  70.76488  70.76488  70.76488
#>  [85]  70.76417  70.71077  70.71074  70.71072  70.71071  70.71069  70.71068
#>  [92]  70.71067  70.71066  70.71065  70.71065  70.71064  70.71063  70.71063
#>  [99]  70.71062  70.71062
#> 
#> $linear.predictors
#>      5.3395   4.9797   4.6441   4.3311   4.0392   3.7669   3.5131   3.2763
#> 1  19.58095 19.90382 20.22628 20.52700 20.80746 21.06901 21.31294 21.54042
#> 2  19.58095 19.51647 19.45207 19.39201 19.33599 19.28376 19.23504 19.18961
#> 3  19.58095 19.11331 18.64626 18.21069 17.80447 17.42564 17.07233 16.74284
#> 4  19.58095 19.11331 18.64626 18.21069 17.80447 17.42564 17.07233 16.74284
#> 5  19.58095 19.95639 20.33135 20.68104 21.00716 21.31130 21.59494 21.85946
#> 6  19.58095 19.97971 20.37796 20.74937 21.09574 21.41877 21.72003 22.00099
#> 7  19.58095 19.87378 20.16624 20.43898 20.69334 20.93056 21.15179 21.35811
#> 8  19.58095 19.44611 19.31145 19.18586 19.06873 18.95950 18.85763 18.76262
#> 9  19.58095 19.44611 19.31145 19.18586 19.06873 18.95950 18.85763 18.76262
#> 10 19.58095 19.44611 19.31145 19.18586 19.06873 18.95950 18.85763 18.76262
#> 11 19.58095 18.67062 17.76145 16.91355 16.12280 15.38535 14.69759 14.05619
#> 12 19.58095 18.71805 17.85625 17.05253 16.30298 15.60395 14.95203 14.34405
#> 13 19.58095 18.79710 18.01425 17.28416 16.60328 15.96829 15.37609 14.82380
#> 14 19.58095 20.22517 20.86856 21.46858 22.02817 22.55004 23.03674 23.49064
#> 15 19.58095 20.25521 20.92860 21.55660 22.14228 22.68849 23.19789 23.67295
#> 16 19.58095 20.06153 20.54149 20.98911 21.40655 21.79586 22.15894 22.49754
#> 17 19.58095 19.27931 18.97806 18.69711 18.43510 18.19074 17.96286 17.75033
#> 18 19.58095 19.15283 18.72526 18.32650 17.95462 17.60780 17.28436 16.98272
#> 19 19.58095 20.22398 20.86619 21.46511 22.02367 22.54458 23.03038 23.48344
#> 20 19.58095 19.96311 20.34478 20.70072 21.03268 21.34226 21.63098 21.90024
#> 21 19.58095 20.05797 20.53438 20.97868 21.39304 21.77947 22.13985 22.47595
#>      3.0555   2.8496   2.6575   2.4784   2.3114   2.1556   2.0103   1.8748
#> 1  21.75258 21.95043 22.13495 22.30704 22.46753 22.61720 22.75678 22.88696
#> 2  19.14724 19.10772 19.07087 19.03650 19.00445 18.97455 18.94668 18.92068
#> 3  16.43555 16.14898 15.88171 15.63246 15.40001 15.18323 14.98106 14.79251
#> 4  16.43555 16.14898 15.88171 15.63246 15.40001 15.18323 14.98106 14.79251
#> 5  22.10616 22.33623 22.55079 22.75090 22.93752 23.11156 23.27387 23.42524
#> 6  22.26301 22.50737 22.73527 22.94780 23.14601 23.33086 23.50325 23.66402
#> 7  21.55053 21.72998 21.89733 22.05341 22.19896 22.33471 22.46130 22.57937
#> 8  18.67402 18.59139 18.51433 18.44246 18.37544 18.31293 18.25464 18.20027
#> 9  18.67402 18.59139 18.51433 18.44246 18.37544 18.31293 18.25464 18.20027
#> 10 18.67402 18.59139 18.51433 18.44246 18.37544 18.31293 18.25464 18.20027
#> 11 13.45802 12.90016 12.37990 11.89470 11.44221 11.02021 10.62665 10.25962
#> 12 13.77704 13.24825 12.75509 12.29518 11.86626 11.46625 11.09319 10.74528
#> 13 14.30874 13.82839 13.38042 12.96263 12.57301 12.20964 11.87077 11.55473
#> 14 23.91395 24.30873 24.67690 25.02025 25.34047 25.63911 25.91761 26.17735
#> 15 24.11600 24.52918 24.91452 25.27389 25.60904 25.92160 26.21309 26.48494
#> 16 22.81332 23.10782 23.38248 23.63862 23.87750 24.10027 24.30804 24.50180
#> 17 17.55213 17.36728 17.19489 17.03413 16.88419 16.74436 16.61396 16.49234
#> 18 16.70140 16.43905 16.19438 15.96619 15.75339 15.55493 15.36984 15.19723
#> 19 23.90597 24.30002 24.66752 25.01024 25.32987 25.62795 25.90595 26.16521
#> 20 22.15135 22.38554 22.60395 22.80763 22.99759 23.17474 23.33996 23.49404
#> 21 22.78940 23.08172 23.35434 23.60858 23.84569 24.06682 24.27305 24.46537
#>       1.7485    1.6306    1.5207    1.4182    1.3226    1.2335    1.1504
#> 1  22.985831 22.985831 22.985831 22.850897 22.589765 22.346534 22.119452
#> 2  18.900929 18.900929 18.900929 19.051197 19.342002 19.612873 19.865759
#> 3  14.649297 14.649297 14.649297 14.760765 14.976484 15.177416 15.365008
#> 4  14.649297 14.649297 14.649297 14.597929 14.498520 14.405925 14.319478
#> 5  23.540211 23.540211 23.540211 23.562556 23.605798 23.646076 23.683680
#> 6  23.786139 23.786139 23.786139 23.806018 23.844490 23.880324 23.913779
#> 7  22.669043 22.669043 22.669043 22.537284 22.282298 22.044791 21.823053
#> 8  18.158978 18.158978 18.158978 18.153847 18.143917 18.134667 18.126032
#> 9  18.158978 18.158978 18.158978 18.153847 18.143917 18.134667 18.126032
#> 10 18.158978 18.158978 18.158978 18.153847 18.143917 18.134667 18.126032
#> 11  9.980837  9.980837  9.980837  9.976267  9.967422  9.959184  9.951493
#> 12 10.481029 10.481029 10.481029 10.471445 10.452897 10.435620 10.419491
#> 13 11.314683 11.314683 11.314683 11.296742 11.262022 11.229681 11.199488
#> 14 26.374633 26.374633 26.374633 26.449983 26.595804 26.731629 26.858436
#> 15 26.691421 26.691421 26.691421 26.763596 26.903271 27.033372 27.154835
#> 16 24.648970 24.648970 24.648970 24.741618 24.920916 25.087923 25.243841
#> 17 16.399969 16.399969 16.399969 16.493888 16.675646 16.844944 17.003002
#> 18 15.066123 15.066123 15.066123 15.010577 14.903082 14.802955 14.709477
#> 19 26.362128 26.362128 26.362128 26.437603 26.583667 26.719718 26.846736
#> 20 23.611072 23.611072 23.611072 23.307034 22.718644 22.170588 21.658920
#> 21 24.611456 24.611456 24.611456 24.623062 24.645523 24.666444 24.685977
#>      1.0728   1.0005    0.9331    0.8702   0.8116   0.7569   0.7059   0.6583
#> 1  21.90767 21.71017 21.525976 21.354196 21.22757 21.40920 21.58844 21.80775
#> 2  20.10160 20.32155 20.526675 20.717975 20.89777 21.08298 21.22083 21.16593
#> 3  15.53996 15.70311 15.855275 15.997181 16.15059 16.48820 16.76482 16.82193
#> 4  14.23886 14.16367 14.093550 14.028156 13.95574 13.78297 13.64053 13.60581
#> 5  23.71875 23.75146 23.781957 23.810403 23.78287 23.26816 22.81640 22.54326
#> 6  23.94498 23.97408 24.001214 24.026521 23.99398 23.45569 22.98379 22.70152
#> 7  21.61626 21.42340 21.243543 21.075806 20.84017 19.89861 19.09915 18.76654
#> 8  18.11798 18.11047 18.103464 18.096931 18.06600 17.81197 17.59153 17.47236
#> 9  18.11798 18.11047 18.103464 18.096931 18.06600 17.81197 17.59153 17.47236
#> 10 18.11798 18.11047 18.103464 18.096931 18.06600 17.81197 17.59153 17.47236
#> 11  9.94432  9.93763  9.931391  9.925573  9.94817 10.22294 10.46295 10.60164
#> 12 10.40445 10.39042 10.377337 10.365136 10.37755 10.60437 10.80341 10.92352
#> 13 11.17133 11.14507 11.120580 11.097740 11.09319 11.24009 11.37083 11.45998
#> 14 26.97670 27.08699 27.189845 27.285770 27.42892 28.05123 28.58145 28.81264
#> 15 27.26811 27.37375 27.472277 27.564159 27.70086 28.29281 28.79707 29.01650
#> 16 25.38925 25.52486 25.651332 25.769279 25.83209 25.46627 25.13322 24.86482
#> 17 17.15041 17.28788 17.416085 17.535650 17.65343 17.82321 17.95641 17.94850
#> 18 14.62230 14.54099 14.465171 14.394458 14.31356 14.10083 13.92424 13.87404
#> 19 26.96519 27.07567 27.178696 27.274780 27.41818 28.04170 28.57294 28.80460
#> 20 21.18174 20.73671 20.321682 19.934623 19.56944 19.18076 18.88972 18.99397
#> 21 24.70419 24.72118 24.737024 24.751799 24.81792 25.35407 25.81921 26.06997
#>      0.6139   0.5725   0.5340   0.4980   0.4644   0.4331   0.4039   0.3767
#> 1  22.01237 22.19033 22.19040 22.19044 22.19046 22.19048 22.19048 22.19049
#> 2  21.11474 21.07025 21.07034 21.07039 21.07042 21.07044 21.07045 21.07045
#> 3  16.87525 16.92167 16.92182 16.92191 16.92196 16.92198 16.92200 16.92201
#> 4  13.57340 13.54519 13.54512 13.54507 13.54505 13.54503 13.54503 13.54502
#> 5  22.28841 22.06669 22.06647 22.06635 22.06628 22.06624 22.06622 22.06620
#> 6  22.43814 22.20901 22.20878 22.20865 22.20857 22.20853 22.20851 22.20849
#> 7  18.45616 18.18609 18.18568 18.18544 18.18531 18.18523 18.18519 18.18517
#> 8  17.36117 17.26443 17.26432 17.26425 17.26422 17.26420 17.26419 17.26418
#> 9  17.36117 17.26443 17.26432 17.26425 17.26422 17.26420 17.26419 17.26418
#> 10 17.36117 17.26443 17.26432 17.26425 17.26422 17.26420 17.26419 17.26418
#> 11 10.73105 10.84363 10.84375 10.84382 10.84386 10.84388 10.84389 10.84390
#> 12 11.03559 11.13309 11.13318 11.13324 11.13327 11.13329 11.13330 11.13330
#> 13 11.54315 11.61551 11.61557 11.61561 11.61563 11.61564 11.61564 11.61565
#> 14 29.02838 29.21609 29.21636 29.21651 29.21660 29.21665 29.21668 29.21670
#> 15 29.22125 29.39941 29.39966 29.39981 29.39990 29.39994 29.39997 29.39999
#> 16 24.61439 24.39655 24.39640 24.39631 24.39627 24.39624 24.39622 24.39621
#> 17 17.94113 17.93475 17.93483 17.93488 17.93490 17.93492 17.93493 17.93493
#> 18 13.82719 13.78641 13.78631 13.78626 13.78623 13.78621 13.78620 13.78619
#> 19 29.02076 29.20885 29.20912 29.20928 29.20937 29.20942 29.20944 29.20946
#> 20 19.09119 19.17568 19.17549 19.17538 19.17532 19.17529 19.17527 19.17526
#> 21 26.30396 26.50753 26.50776 26.50789 26.50796 26.50801 26.50803 26.50804
#>      0.3513   0.3276   0.3055   0.2850   0.2658   0.2478   0.2311   0.2156
#> 1  21.10983 21.10956 21.10934 21.10915 21.10900 21.10886 21.10875 21.10866
#> 2  21.27398 21.27414 21.27426 21.27437 21.27446 21.27454 21.27460 21.27466
#> 3  17.70575 17.70607 17.70633 17.70655 17.70674 17.70689 17.70703 17.70714
#> 4  12.81021 12.81007 12.80995 12.80985 12.80976 12.80969 12.80963 12.80958
#> 5  23.70749 23.70766 23.70781 23.70793 23.70804 23.70812 23.70820 23.70826
#> 6  23.77230 23.77245 23.77258 23.77268 23.77278 23.77285 23.77292 23.77297
#> 7  18.58236 18.58216 18.58199 18.58185 18.58173 18.58163 18.58154 18.58147
#> 8  16.18291 16.18268 16.18248 16.18231 16.18217 16.18205 16.18195 16.18187
#> 9  16.18291 16.18268 16.18248 16.18231 16.18217 16.18205 16.18195 16.18187
#> 10 16.18291 16.18268 16.18248 16.18231 16.18217 16.18205 16.18195 16.18187
#> 11 11.57990 11.58019 11.58042 11.58062 11.58079 11.58094 11.58106 11.58116
#> 12 11.71172 11.71196 11.71216 11.71233 11.71247 11.71259 11.71269 11.71278
#> 13 11.93142 11.93158 11.93172 11.93183 11.93193 11.93201 11.93208 11.93214
#> 14 29.34621 29.34632 29.34641 29.34649 29.34656 29.34661 29.34666 29.34670
#> 15 29.42970 29.42978 29.42985 29.42990 29.42995 29.42999 29.43003 29.43005
#> 16 22.78880 22.78843 22.78812 22.78785 22.78763 22.78744 22.78728 22.78715
#> 17 18.16712 18.16727 18.16740 18.16751 18.16760 18.16768 18.16774 18.16780
#> 18 12.92006 12.91988 12.91973 12.91960 12.91949 12.91940 12.91932 12.91926
#> 19 29.34292 29.34303 29.34312 29.34320 29.34326 29.34332 29.34337 29.34341
#> 20 20.03771 20.03761 20.03752 20.03745 20.03739 20.03734 20.03729 20.03726
#> 21 26.43378 26.43382 26.43386 26.43389 26.43391 26.43394 26.43395 26.43397
#>      0.2010   0.1875   0.1748   0.1631   0.1521   0.1418   0.1323   0.1234
#> 1  21.15999 21.15973 21.04795 21.04772 21.04754 21.04740 21.04730 21.04724
#> 2  20.81413 20.81441 20.40383 20.40416 20.40446 20.40471 20.40494 20.40513
#> 3  17.70601 17.70630 18.93136 18.93136 18.93131 18.93122 18.93109 18.93094
#> 4  12.92952 12.92934 12.71502 12.71491 12.71482 12.71476 12.71473 12.71470
#> 5  23.66517 23.66534 23.18993 23.19022 23.19046 23.19068 23.19086 23.19101
#> 6  24.10084 24.10090 24.04133 24.04143 24.04150 24.04155 24.04158 24.04160
#> 7  19.01562 19.01530 19.78934 19.78888 19.78846 19.78809 19.78776 19.78747
#> 8  15.78145 15.78133 16.19800 16.19775 16.19754 16.19736 16.19722 16.19710
#> 9  15.78145 15.78133 16.19800 16.19775 16.19754 16.19736 16.19722 16.19710
#> 10 15.78145 15.78133 16.19800 16.19775 16.19754 16.19736 16.19722 16.19710
#> 11 11.58965 11.58990 10.73661 10.73707 10.73748 10.73783 10.73814 10.73840
#> 12 11.80105 11.80124 11.08743 11.08780 11.08813 11.08841 11.08865 11.08887
#> 13 12.34561 12.34565 12.06554 12.06567 12.06579 12.06588 12.06595 12.06601
#> 14 29.41054 29.41065 29.62682 29.62687 29.62691 29.62693 29.62694 29.62694
#> 15 29.70669 29.70672 30.18109 30.18101 30.18093 30.18084 30.18076 30.18067
#> 16 22.93070 22.93034 21.38174 21.38174 21.38181 21.38193 21.38209 21.38228
#> 17 17.36888 17.36923 17.95501 17.95516 17.95527 17.95535 17.95539 17.95541
#> 18 13.88873 13.88831 14.60992 14.60939 14.60891 14.60850 14.60813 14.60781
#> 19 29.40822 29.40834 29.62412 29.62418 29.62421 29.62424 29.62425 29.62424
#> 20 19.31865 19.31870 19.08044 19.08052 19.08060 19.08068 19.08076 19.08084
#> 21 26.69562 26.69561 26.13853 26.13866 26.13879 26.13892 26.13904 26.13915
#>      0.1150   0.1073   0.1001   0.0933   0.0870   0.0812   0.0757   0.0706
#> 1  21.04720 21.04719 21.04719 21.12274 21.12260 21.12246 21.33392 21.33405
#> 2  20.40529 20.40544 20.40556 20.32826 20.32825 20.32825 20.53614 20.53602
#> 3  18.93077 18.93058 18.93039 18.68947 18.68987 18.69025 19.05415 19.05370
#> 4  12.71470 12.71471 12.71472 12.47613 12.47597 12.47583 13.13062 13.13051
#> 5  23.19114 23.19125 23.19133 22.84878 22.84869 22.84862 22.47307 22.47308
#> 6  24.04160 24.04160 24.04159 24.54837 24.54849 24.54859 25.01191 25.01176
#> 7  19.78721 19.78699 19.78679 19.78587 19.78586 19.78585 19.48560 19.48573
#> 8  16.19700 16.19692 16.19686 16.35080 16.35077 16.35074 15.79191 15.79217
#> 9  16.19700 16.19692 16.19686 16.40753 16.40751 16.40749 16.33023 16.33037
#> 10 16.19700 16.19692 16.19686 16.52099 16.52099 16.52098 16.50174 16.50187
#> 11 10.73863 10.73882 10.73898 10.91660 10.91666 10.91672 11.06873 11.06861
#> 12 11.08905 11.08920 11.08933 11.24512 11.24517 11.24522 11.02712 11.02712
#> 13 12.06606 12.06609 12.06612 12.15649 12.15656 12.15663 11.72270 11.72276
#> 14 29.62693 29.62691 29.62689 29.68920 29.68932 29.68944 29.47827 29.47826
#> 15 30.18058 30.18049 30.18041 30.38703 30.38721 30.38736 30.67648 30.67634
#> 16 21.38250 21.38274 21.38298 21.21849 21.21789 21.21735 21.38434 21.38468
#> 17 17.95541 17.95539 17.95535 17.70502 17.70519 17.70536 17.64537 17.64519
#> 18 14.60752 14.60728 14.60706 14.30196 14.30195 14.30193 14.33702 14.33702
#> 19 29.62423 29.62422 29.62420 29.52437 29.52447 29.52456 29.47336 29.47330
#> 20 19.08092 19.08100 19.08108 18.91712 18.91700 18.91690 19.12719 19.12717
#> 21 26.13925 26.13935 26.13945 26.05965 26.05956 26.05948 25.61013 25.61030
#>      0.0658   0.0614   0.0573   0.0534   0.0498   0.0464   0.0433   0.0404
#> 1  21.33417 21.33427 21.33436 21.33445 21.33452 21.33459 21.33465 21.33471
#> 2  20.53589 20.53577 20.53565 20.53553 20.53541 20.53529 20.53517 20.53506
#> 3  19.05328 19.05288 19.05251 19.05217 19.05185 19.05156 19.05129 19.05104
#> 4  13.13040 13.13028 13.13015 13.13002 13.12989 13.12976 13.12962 13.12949
#> 5  22.47309 22.47310 22.47311 22.47312 22.47313 22.47314 22.47316 22.47317
#> 6  25.01161 25.01148 25.01136 25.01124 25.01114 25.01104 25.01095 25.01087
#> 7  19.48586 19.48600 19.48615 19.48629 19.48644 19.48659 19.48674 19.48689
#> 8  15.79243 15.79268 15.79293 15.79317 15.79341 15.79363 15.79386 15.79407
#> 9  16.33050 16.33062 16.33075 16.33087 16.33099 16.33110 16.33121 16.33132
#> 10 16.50200 16.50212 16.50224 16.50236 16.50248 16.50259 16.50270 16.50281
#> 11 11.06849 11.06837 11.06825 11.06812 11.06800 11.06788 11.06775 11.06763
#> 12 11.02712 11.02711 11.02710 11.02709 11.02708 11.02706 11.02704 11.02702
#> 13 11.72283 11.72289 11.72295 11.72301 11.72307 11.72312 11.72318 11.72323
#> 14 29.47826 29.47827 29.47828 29.47829 29.47831 29.47833 29.47835 29.47838
#> 15 30.67622 30.67610 30.67600 30.67590 30.67581 30.67573 30.67566 30.67560
#> 16 21.38498 21.38525 21.38549 21.38570 21.38589 21.38605 21.38618 21.38630
#> 17 17.64501 17.64485 17.64470 17.64456 17.64443 17.64432 17.64421 17.64411
#> 18 14.33703 14.33704 14.33705 14.33707 14.33709 14.33711 14.33713 14.33715
#> 19 29.47323 29.47317 29.47312 29.47307 29.47303 29.47299 29.47296 29.47292
#> 20 19.12715 19.12713 19.12710 19.12707 19.12703 19.12699 19.12695 19.12691
#> 21 25.61046 25.61061 25.61075 25.61088 25.61101 25.61112 25.61123 25.61133
#>      0.0377   0.0351   0.0328   0.0306   0.0285   0.0266   0.0248   0.0231
#> 1  21.33475 21.33517 21.75512 21.75509 21.75507 21.75505 21.75503 21.75502
#> 2  20.53495 20.53476 20.59617 20.59618 20.59619 20.59620 20.59621 20.59622
#> 3  19.05081 19.05034 19.09454 19.09431 19.09411 19.09392 19.09375 19.09360
#> 4  13.12935 13.12886 12.98761 12.98778 12.98793 12.98806 12.98818 12.98827
#> 5  22.47318 22.47363 22.59128 22.59139 22.59151 22.59163 22.59176 22.59189
#> 6  25.01080 25.01055 24.89031 24.89021 24.89011 24.89002 24.88993 24.88984
#> 7  19.48704 19.48785 19.67864 19.67851 19.67841 19.67832 19.67825 19.67819
#> 8  15.79428 15.79472 15.68718 15.68707 15.68697 15.68687 15.68679 15.68671
#> 9  16.33142 16.33164 16.28021 16.28008 16.27995 16.27983 16.27972 16.27961
#> 10 16.50291 16.50309 16.44499 16.44483 16.44467 16.44453 16.44439 16.44425
#> 11 11.06751 11.06727 11.23254 11.23259 11.23264 11.23268 11.23272 11.23275
#> 12 11.02700 11.02688 11.06034 11.06041 11.06048 11.06054 11.06059 11.06064
#> 13 11.72328 11.72321 11.63722 11.63729 11.63735 11.63741 11.63746 11.63751
#> 14 29.47840 29.47829 29.42494 29.42489 29.42484 29.42479 29.42475 29.42471
#> 15 30.67554 30.67523 30.70461 30.70447 30.70434 30.70422 30.70411 30.70400
#> 16 21.38640 21.38653 21.37339 21.37377 21.37412 21.37444 21.37473 21.37499
#> 17 17.64402 17.64402 17.70193 17.70178 17.70164 17.70151 17.70141 17.70131
#> 18 14.33717 14.33708 14.24368 14.24375 14.24382 14.24388 14.24394 14.24399
#> 19 29.47290 29.47272 29.46120 29.46118 29.46116 29.46114 29.46113 29.46112
#> 20 19.12687 19.12687 18.93845 18.93857 18.93868 18.93877 18.93885 18.93892
#> 21 25.61142 25.61131 25.41565 25.41584 25.41602 25.41618 25.41632 25.41646
#>      0.0216   0.0201   0.0187   0.0175   0.0163   0.0152   0.0142   0.0132
#> 1  21.75500 21.75500 21.75499 21.75499 21.75499 21.75394 21.69278 21.69243
#> 2  20.59623 20.59624 20.59625 20.59626 20.59626 20.59713 20.72223 20.72257
#> 3  19.09346 19.09334 19.09323 19.09313 19.09303 19.09206 19.03697 19.03674
#> 4  12.98835 12.98842 12.98848 12.98852 12.98856 12.98829 13.02007 13.02063
#> 5  22.59201 22.59214 22.59226 22.59238 22.59249 22.59358 22.52400 22.52335
#> 6  24.88975 24.88967 24.88959 24.88951 24.88944 24.89000 24.83344 24.83336
#> 7  19.67815 19.67811 19.67809 19.67808 19.67807 19.67970 19.81297 19.81270
#> 8  15.68664 15.68658 15.68652 15.68647 15.68642 15.68592 15.67127 15.67116
#> 9  16.27951 16.27942 16.27933 16.27925 16.27918 16.27852 16.27810 16.27815
#> 10 16.44413 16.44401 16.44390 16.44380 16.44371 16.44300 16.44480 16.44489
#> 11 11.23278 11.23281 11.23283 11.23285 11.23287 11.23300 11.24718 11.24727
#> 12 11.06068 11.06072 11.06076 11.06079 11.06082 11.06102 11.06200 11.06205
#> 13 11.63755 11.63759 11.63763 11.63766 11.63769 11.63787 11.61916 11.61913
#> 14 29.42467 29.42464 29.42461 29.42458 29.42456 29.42455 29.43210 29.43211
#> 15 30.70389 30.70380 30.70371 30.70362 30.70355 30.70332 30.71647 30.71655
#> 16 21.37524 21.37545 21.37565 21.37583 21.37600 21.37670 21.34127 21.34139
#> 17 17.70123 17.70116 17.70109 17.70104 17.70099 17.70015 17.67211 17.67191
#> 18 14.24404 14.24409 14.24413 14.24416 14.24420 14.24403 14.21802 14.21799
#> 19 29.46111 29.46110 29.46110 29.46109 29.46109 29.46107 29.47243 29.47245
#> 20 18.93898 18.93903 18.93907 18.93911 18.93914 18.93860 18.93500 18.93532
#> 21 25.41658 25.41669 25.41679 25.41687 25.41695 25.41753 25.44763 25.44788
#>      0.0123   0.0115   0.0107   0.0100   0.0093   0.0087   0.0081   0.0076
#> 1  21.69210 21.69178 21.69147 21.69118 21.69090 21.69064 21.69040 21.69016
#> 2  20.72287 20.72315 20.72340 20.72362 20.72382 20.72399 20.72415 20.72428
#> 3  19.03653 19.03634 19.03617 19.03600 19.03586 19.03573 19.03561 19.03550
#> 4  13.02115 13.02164 13.02210 13.02253 13.02294 13.02332 13.02367 13.02400
#> 5  22.52275 22.52221 22.52172 22.52127 22.52086 22.52050 22.52018 22.51989
#> 6  24.83328 24.83320 24.83314 24.83308 24.83302 24.83297 24.83293 24.83289
#> 7  19.81242 19.81214 19.81186 19.81158 19.81131 19.81103 19.81076 19.81050
#> 8  15.67106 15.67097 15.67087 15.67079 15.67070 15.67062 15.67054 15.67046
#> 9  16.27819 16.27821 16.27823 16.27825 16.27825 16.27825 16.27824 16.27822
#> 10 16.44496 16.44502 16.44507 16.44510 16.44513 16.44514 16.44515 16.44514
#> 11 11.24736 11.24743 11.24751 11.24757 11.24763 11.24768 11.24772 11.24776
#> 12 11.06209 11.06214 11.06218 11.06223 11.06227 11.06231 11.06235 11.06239
#> 13 11.61912 11.61911 11.61911 11.61911 11.61912 11.61913 11.61915 11.61917
#> 14 29.43211 29.43211 29.43212 29.43212 29.43212 29.43212 29.43212 29.43212
#> 15 30.71661 30.71666 30.71670 30.71674 30.71676 30.71678 30.71679 30.71679
#> 16 21.34151 21.34164 21.34176 21.34189 21.34202 21.34215 21.34228 21.34240
#> 17 17.67172 17.67155 17.67138 17.67123 17.67110 17.67097 17.67086 17.67075
#> 18 14.21797 14.21795 14.21795 14.21795 14.21795 14.21796 14.21798 14.21800
#> 19 29.47246 29.47248 29.47250 29.47252 29.47253 29.47255 29.47256 29.47258
#> 20 18.93563 18.93592 18.93621 18.93648 18.93673 18.93698 18.93722 18.93744
#> 21 25.44811 25.44834 25.44856 25.44878 25.44898 25.44918 25.44937 25.44955
#>      0.0071   0.0066   0.0061   0.0057   0.0053
#> 1  21.68994 21.68974 21.68954 21.68936 21.68919
#> 2  20.72439 20.72449 20.72457 20.72463 20.72468
#> 3  19.03540 19.03532 19.03524 19.03517 19.03512
#> 4  13.02431 13.02459 13.02486 13.02510 13.02532
#> 5  22.51964 22.51942 22.51923 22.51906 22.51892
#> 6  24.83285 24.83282 24.83279 24.83277 24.83274
#> 7  19.81023 19.80998 19.80973 19.80948 19.80925
#> 8  15.67039 15.67032 15.67025 15.67019 15.67013
#> 9  16.27821 16.27818 16.27815 16.27812 16.27809
#> 10 16.44513 16.44511 16.44509 16.44506 16.44503
#> 11 11.24780 11.24783 11.24786 11.24788 11.24790
#> 12 11.06243 11.06246 11.06250 11.06253 11.06256
#> 13 11.61919 11.61922 11.61925 11.61928 11.61932
#> 14 29.43213 29.43213 29.43213 29.43213 29.43213
#> 15 30.71679 30.71679 30.71677 30.71676 30.71674
#> 16 21.34253 21.34265 21.34278 21.34290 21.34301
#> 17 17.67066 17.67057 17.67049 17.67042 17.67036
#> 18 14.21803 14.21805 14.21808 14.21812 14.21815
#> 19 29.47260 29.47261 29.47263 29.47264 29.47265
#> 20 18.93765 18.93785 18.93804 18.93822 18.93839
#> 21 25.44972 25.44988 25.45004 25.45019 25.45033
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.4 18.7 14.3 24.4 22.8 19.2 16.4 17.3 15.2 10.4 10.4 14.7 32.4 33.9
#> [16] 21.5 15.5 13.3 27.3 19.7 21.4
#> 
#> $X
#>               am       carb        cyl       disp       drat       gear
#>  [1,]  1.5811388  0.8992288 -0.1608169 -0.6405304  0.8744320  0.8922178
#>  [2,] -0.6324555 -1.2796718 -0.1608169  0.1279269 -0.9267306 -0.8111071
#>  [3,] -0.6324555 -0.5533716  0.9649013  0.9277498 -0.7729728 -0.8111071
#>  [4,] -0.6324555  0.8992288  0.9649013  0.9277498 -0.6411804 -0.8111071
#>  [5,] -0.6324555 -0.5533716 -1.2865350 -0.7448211  0.4131587  0.8922178
#>  [6,] -0.6324555 -0.5533716 -1.2865350 -0.7910853  0.9183628  0.8922178
#>  [7,] -0.6324555  0.8992288 -0.1608169 -0.5809358  0.9183628  0.8922178
#>  [8,] -0.6324555  0.1729286  0.9649013  0.2675038 -0.9486960 -0.8111071
#>  [9,] -0.6324555  0.1729286  0.9649013  0.2675038 -0.9486960 -0.8111071
#> [10,] -0.6324555  0.1729286  0.9649013  0.2675038 -0.9486960 -0.8111071
#> [11,] -0.6324555  0.8992288  0.9649013  1.8059867 -1.2562116 -0.8111071
#> [12,] -0.6324555  0.8992288  0.9649013  1.7118899 -1.1024538 -0.8111071
#> [13,] -0.6324555  0.8992288  0.9649013  1.5550618 -0.5972496 -0.8111071
#> [14,]  1.5811388 -1.2796718 -1.2865350 -1.2780363  1.2698092  0.8922178
#> [15,]  1.5811388 -1.2796718 -1.2865350 -1.3376310  1.5773247  0.8922178
#> [16,] -0.6324555 -1.2796718 -1.2865350 -0.9534023  0.4351241 -0.8111071
#> [17,] -0.6324555 -0.5533716  0.9649013  0.5984109 -1.6296233 -0.8111071
#> [18,] -0.6324555  0.8992288  0.9649013  0.8493357  0.5010203 -0.8111071
#> [19,]  1.5811388 -1.2796718 -1.2865350 -1.2756839  1.2698092  0.8922178
#> [20,]  1.5811388  2.3518292 -0.1608169 -0.7581515  0.2594009  2.5955427
#> [21,]  1.5811388 -0.5533716 -1.2865350 -0.9463451  1.3357054  0.8922178
#>                hp        qsec         vs          wt
#>  [1,] -0.61203158 -0.96042144 -0.8660254 -0.79405907
#>  [2,] -0.61203158  0.74802579  1.1547005 -0.19452024
#>  [3,]  0.46872276 -0.63937095 -0.8660254  0.03219612
#>  [4,]  1.63261204 -1.31587019 -0.8660254  0.16318780
#>  [5,] -1.41012709  1.06907628  1.1547005 -0.21971095
#>  [6,] -0.86143642  2.73165916  1.1547005 -0.26001608
#>  [7,] -0.39588071  0.09445873  1.1547005  0.03219612
#>  [8,]  0.55185771 -0.42151526 -0.8660254  0.66700194
#>  [9,]  0.55185771 -0.30685438 -0.8660254  0.32440833
#> [10,]  0.55185771 -0.07753260 -0.8660254  0.37478974
#> [11,]  0.96753245 -0.08899869 -0.8660254  1.85600333
#> [12,]  1.13380235 -0.18072740 -0.8660254  2.03133065
#> [13,]  1.38320720 -0.41004918 -0.8660254  1.95172801
#> [14,] -1.34361913  0.76522493  1.1547005 -1.21726296
#> [15,] -1.36024612  1.01174584  1.1547005 -1.58504728
#> [16,] -0.82818244  1.07480932  1.1547005 -0.95024146
#> [17,]  0.05304802 -0.72536662 -0.8660254  0.11280639
#> [18,]  1.63261204 -1.56239110 -0.8660254  0.43524744
#> [19,] -1.34361913  0.43844140  1.1547005 -1.48428445
#> [20,]  0.46872276 -1.51079370 -0.8660254 -0.64291483
#> [21,] -0.62865857  0.26645006  1.1547005 -0.63283855
#> attr(,"center")
#>  [1]   0.2857143   2.7619048   6.2857143 241.6857143   3.5019048   3.4761905
#>  [7] 146.8095238  18.1352381   0.4285714   3.4080476
#> attr(,"scale")
#>  [1]   0.4517540   1.3768412   1.7766436 127.5282314   0.4552615   0.5870870
#>  [7]  60.1431776   1.7442739   0.4948717   0.9924295
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 34
#> 
#> $lambda.min
#> [1] 0.5339541
#> 
#> $null.dev
#> [1] 37.46154
#> 
#> $Bias
#> [1] 5.361001
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 9.005778