Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 48.33315 43.55495 37.56783 32.56439 28.40276 24.96049 22.13193 19.82599
#>   [9] 17.96422 16.47907 15.31244 14.41442 13.74215 13.25890 12.93322 12.73820
#>  [17] 12.44854 12.35890 12.35890 12.35890 12.35890 12.35890 12.34328 12.19796
#>  [25] 12.04585 11.76007 11.50639 11.28103 11.08063 10.93694 10.82677 12.03093
#>  [33] 12.05029 11.98502 12.00593 11.79334 11.92188 11.75419 12.16038 12.33826
#>  [41] 12.20362 12.21313 12.21337 11.86396 11.69250 11.65933 11.95287 12.99932
#>  [49] 13.13154 13.13196 13.13227 13.02695 12.62834 12.46000 12.63574 13.01132
#>  [57] 15.43487 15.66730 15.91451 15.97236 14.75942 16.25057 16.38122 16.26996
#>  [65] 16.55595 16.55506 16.55422 16.51184 16.68855 16.75764 17.26375 17.25078
#>  [73] 17.75340 17.76581 17.76592 17.68085 17.68087 17.68089 17.99797 17.99825
#>  [81] 17.99854 17.99881 17.99909 17.99936 18.27015 18.27040 18.29751 18.31268
#>  [89] 18.31284 18.31300 18.31315 18.31330 18.33068 18.33066 18.33064 18.33062
#>  [97] 18.32407 18.32402 18.31232 18.24819
#> 
#> $cvse
#>   [1] 12.296230 11.770984 10.587908  9.522228  8.559396  7.687945  6.899057
#>   [8]  6.186223  5.544985  4.972742  4.468580  4.033048  3.667785  3.374826
#>  [15]  3.155542  3.009306  2.895603  2.847687  2.847687  2.847687  2.847687
#>  [22]  2.847687  2.846354  2.836333  2.823362  2.770413  2.730380  2.700514
#>  [29]  2.678589  2.661310  2.647920  2.909293  2.895003  3.072068  3.107694
#>  [36]  3.065255  2.903200  2.950026  3.219633  3.205415  3.142108  3.138728
#>  [43]  3.138783  3.145608  3.061119  3.062784  3.171306  3.233007  3.242222
#>  [50]  3.242308  3.242367  3.239230  3.089208  2.928117  3.044645  3.079560
#>  [57]  4.368388  4.417646  4.422577  4.435001  4.382914  5.384701  5.387760
#>  [64]  4.676668  4.683684  4.683371  4.683062  4.689201  4.673838  4.697059
#>  [71]  4.653336  4.655989  4.729643  4.746664  4.746353  4.705422  4.705088
#>  [78]  4.704759  4.758240  4.758000  4.757765  4.757534  4.757307  4.757084
#>  [85]  4.730769  4.730556  4.727429  4.726246  4.726039  4.725837  4.725639
#>  [92]  4.725446  4.725064  4.724891  4.724722  4.724557  4.725603  4.725451
#>  [99]  4.727506  4.708345
#> 
#> $fold
#>  [1]  9  5  7  8  9  1  3  1  4  7  2  3  5 10 10  1  6  8  4  2  6
#> 
#> $lambda
#>   [1] 6.134459296 5.721017271 5.335439854 4.975849065 4.640493492 4.327739762
#>   [7] 4.036064586 3.764047341 3.510363148 3.273776421 3.053134847 2.847363777
#>  [13] 2.655460988 2.476491805 2.309584546 2.153926277 2.008758855 1.873375231
#>  [19] 1.747116011 1.629366239 1.519552408 1.417139661 1.321629191 1.232555807
#>  [25] 1.149485671 1.072014184 0.999764016 0.932383267 0.869543756 0.810939417
#>  [31] 0.756284814 0.705313749 0.657777963 0.613445930 0.572101727 0.533543985
#>  [37] 0.497584907 0.464049349 0.432773976 0.403606459 0.376404734 0.351036315
#>  [43] 0.327377642 0.305313485 0.284736378 0.265546099 0.247649180 0.230958455
#>  [49] 0.215392628 0.200875886 0.187337523 0.174711601 0.162936624 0.151955241
#>  [55] 0.141713966 0.132162919 0.123255581 0.114948567 0.107201418 0.099976402
#>  [61] 0.093238327 0.086954376 0.081093942 0.075628481 0.070531375 0.065777796
#>  [67] 0.061344593 0.057210173 0.053354399 0.049758491 0.046404935 0.043277398
#>  [73] 0.040360646 0.037640473 0.035103631 0.032737764 0.030531348 0.028473638
#>  [79] 0.026554610 0.024764918 0.023095845 0.021539263 0.020087589 0.018733752
#>  [85] 0.017471160 0.016293662 0.015195524 0.014171397 0.013216292 0.012325558
#>  [91] 0.011494857 0.010720142 0.009997640 0.009323833 0.008695438 0.008109394
#>  [97] 0.007562848 0.007053137 0.006577780 0.006134459
#> 
#> $fit
#> $beta
#>               6.1345       5.7210       5.3354      4.9758      4.6405
#> (Intercept) 20.23333 21.304318232 22.303122333 23.23461038 24.10331925
#> am           0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> carb         0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> cyl          0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> disp         0.00000 -0.004641273 -0.008969741 -0.01300648 -0.01677116
#> drat         0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> gear         0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> hp           0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> qsec         0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> vs           0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> wt           0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#>                  4.3277      4.0361      3.7640      3.5104      3.2738
#> (Intercept) 24.91348005 25.66903872 26.37367527 27.03082167 27.64367861
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.02028212 -0.02355644 -0.02661009 -0.02945793 -0.03211384
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  3.0531      2.8474      2.6555      2.4765     2.3096
#> (Intercept) 28.21523104 28.74826275 29.24536990 29.70897370 30.1413321
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> disp        -0.03459075 -0.03690072 -0.03905501 -0.04106411 -0.0429378
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#>                  2.1539      2.0088      1.8734      1.7471      1.6294
#> (Intercept) 30.54455109 30.82718183 30.82718183 30.82718183 30.82718183
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.04468521 -0.04591003 -0.04591003 -0.04591003 -0.04591003
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  1.5196      1.4171      1.3216      1.2326      1.1495
#> (Intercept) 30.82718183 30.82718183 30.82718183 30.82718183 30.82718183
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.04591003 -0.04591003 -0.04591003 -0.04591003 -0.04591003
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  1.0720      0.9998      0.9324      0.8695      0.8109
#> (Intercept) 30.82718183 30.82718183 30.82718183 30.82718183 30.82718183
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.04591003 -0.04591003 -0.04591003 -0.04591003 -0.04591003
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.7563      0.7053      0.6578      0.6134      0.5721
#> (Intercept) 30.82718183 30.82718183 30.85484398 28.33886218 25.24640607
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000 -0.02358357 -0.15968456 -0.30832246
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.04591003 -0.04591003 -0.04573790 -0.04319931 -0.04022475
#> drat         0.00000000  0.00000000  0.00000000  0.64256711  1.42691532
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.5335      0.4976      0.4640      0.4328      0.4036
#> (Intercept) 22.36687151 19.67776336 18.83909474 28.53530136 28.53577760
#> am           0.00000000  0.00000000  0.00000000  3.62060911  3.61990420
#> carb        -0.44668881 -0.57593531 -0.64918079 -1.03254371 -1.03236828
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.03745575 -0.03486926 -0.03379878 -0.03066466 -0.03066744
#> drat         2.15727829  2.83932654  3.06124787  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                 0.3764      0.3510      0.3274      0.3053      0.2847
#> (Intercept) 28.5360963 19.05556676 17.09432264 17.09508742 17.09583189
#> am           3.6194325  3.39611174  3.15688235  3.15771762  3.15832864
#> carb        -1.0322509 -0.76282381 -0.75115976 -0.75129166 -0.75139222
#> cyl          0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.0306693 -0.02210451 -0.02129439 -0.02129221 -0.02129086
#> drat         0.0000000  1.57842716  2.08436900  2.08400305  2.08370862
#> gear         0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.0000000  2.66960753  2.73711461  2.73718909  2.73719362
#> wt           0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.2655      0.2476      0.2310      0.2154      0.2009
#> (Intercept) 17.09650816 17.09709410 17.09758400 17.09798189 17.09829707
#> am           3.15876534  3.15906945  3.15927484  3.15940833  3.15949068
#> carb        -0.75146728 -0.75152210 -0.75156120 -0.75158838 -0.75160670
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.02129008 -0.02128968 -0.02128954 -0.02128955 -0.02128965
#> drat         2.08347737  2.08329973  2.08316617  2.08306790  2.08299722
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           2.73715789  2.73710192  2.73703866  2.73697604  2.73691854
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.1873      0.1747      0.1629      0.1520      0.1417
#> (Intercept) 17.09854115 17.09872612 13.96806015 13.96763941 16.31384663
#> am           3.15953761  3.15956077  3.42757157  3.42825927  3.15428859
#> carb        -0.75161857 -0.75162586  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.02128979 -0.02128994 -0.01043051 -0.01042704  0.00000000
#> drat         2.08294764  2.08291386  2.63825966  2.63808590  2.57317574
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000 -0.02653494 -0.02653670 -0.02950483
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           2.73686834  2.73682611  3.57633310  3.57676853  3.72268756
#> wt           0.00000000  0.00000000  0.00000000  0.00000000 -1.25449014
#>                  0.1322      0.1233      0.1149     0.1072      0.1000
#> (Intercept) 16.31656884 16.31805636 16.31882244 16.3191760 16.31930229
#> am           3.15453643  3.15475040  3.15493025  3.1550760  3.15519015
#> carb         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> drat         2.57269368  2.57241049  2.57224557  2.5721518  2.57210079
#> gear         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> hp          -0.02950872 -0.02951153 -0.02951348 -0.0295148 -0.02951566
#> qsec         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> vs           3.72204265  3.72167694  3.72148153  3.7213864  3.72134774
#> wt          -1.25456427 -1.25455964 -1.25452132 -1.2544727 -1.25442516
#>                  0.0932      0.0870      0.0811      0.0756     0.0705
#> (Intercept) 16.31931111 16.31926560 16.42191715 16.42419050 16.4257680
#> am           3.15527695  3.15534135  2.88547680  2.88524969  2.8851434
#> carb         0.00000000  0.00000000  0.21990092  0.21980735  0.2197056
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> drat         2.57207490  2.57206342  2.58956817  2.58926351  2.5890518
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> hp          -0.02951623 -0.02951659 -0.03297579 -0.03297302 -0.0329707
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> vs           3.72133926  3.72134540  3.85624017  3.85582283  3.8554822
#> wt          -1.25438370 -1.25434981 -1.31708677 -1.31741262 -1.3176251
#>                 0.0658      0.0613      0.0572      0.0534      0.0498
#> (Intercept) 16.4268223 16.42749195 16.42788594 16.42808821 16.42816203
#> am           2.8851168  2.88513977  2.88519079  2.88525513  2.88532301
#> carb         0.2196061  0.21951483  0.21943475  0.21936694  0.21931116
#> cyl          0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         2.5889099  2.58881949  2.58876592  2.58873803  2.58872738
#> gear         0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.0329688 -0.03296731 -0.03296614 -0.03296527 -0.03296462
#> qsec         0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           3.8552105  3.85499871  3.85483717  3.85471680  3.85462933
#> wt          -1.3177552 -1.31782713 -1.31785928 -1.31786543 -1.31785560
#>                  0.0464      0.0433      0.0404      0.0376      0.0351
#> (Intercept) 16.42815401 16.42809764 16.42801626 16.42792556 16.42783554
#> am           2.88538834  2.88544765  2.88549926  2.88554271  2.88557827
#> carb         0.21926644  0.21923144  0.21920469  0.21918472  0.21917020
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         2.58872776  2.58873470  2.58874507  2.58875678  2.58876847
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.03296416 -0.03296384 -0.03296362 -0.03296349 -0.03296341
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           3.85456758  3.85452547  3.85449803  3.85448126  3.85447204
#> wt          -1.31783689 -1.31781414 -1.31779056 -1.31776809 -1.31774785
#>                  0.0327      0.0305      0.0285      0.0266      0.0248
#> (Intercept) 15.93683531 15.93686057 15.93688540 15.93690982 15.93693382
#> am           2.67807631  2.67801289  2.67795054  2.67788925  2.67782900
#> carb         0.72083746  0.72099687  0.72115357  0.72130762  0.72145905
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.01544069  0.01544560  0.01545042  0.01545516  0.01545983
#> drat         2.92963660  2.92970577  2.92977377  2.92984061  2.92990632
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.04895318 -0.04895812 -0.04896297 -0.04896773 -0.04897242
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           4.82176919  4.82207194  4.82236955  4.82266211  4.82294971
#> wt          -2.45660282 -2.45698399 -2.45735870 -2.45772705 -2.45808915
#>                  0.0231      0.0215        0.0201      0.0187      0.0175
#> (Intercept) 14.22043602 14.21748795 14.2138261499 16.89957173 16.90193502
#> am           2.80732312  2.80768284  2.8088362386  3.03373971  3.03402848
#> carb         0.75095000  0.75096230  0.7510100412  0.83799983  0.83795605
#> cyl          0.00000000  0.00000000  0.0000000000 -0.29580610 -0.29615538
#> disp         0.01710881  0.01711273  0.0171196217  0.02087384  0.02087552
#> drat         2.85958124  2.85933429  2.8589030242  2.68191456  2.68152836
#> gear         0.00000000  0.00000000 -0.0005069385 -0.27463365 -0.27480865
#> hp          -0.04838680 -0.04838454 -0.0483801000 -0.04747506 -0.04746982
#> qsec         0.12599281  0.12624188  0.1266957400  0.15647466  0.15655953
#> vs           4.57384189  4.57334368  4.5724721416  4.29015576  4.28945889
#> wt          -2.70298081 -2.70355623 -2.7046551799 -3.01690458 -3.01709658
#>                  0.0163      0.0152      0.0142      0.0132      0.0123
#> (Intercept) 16.90423941 16.90648610 16.90867623 16.91081095 16.91289139
#> am           3.03430944  3.03458274  3.03484854  3.03510700  3.03535829
#> carb         0.83791069  0.83786381  0.83781548  0.83776577  0.83771474
#> cyl         -0.29649523 -0.29682585 -0.29714743 -0.29746016 -0.29776424
#> disp         0.02087708  0.02087852  0.02087984  0.02088104  0.02088214
#> drat         2.68115119  2.68078288  2.68042325  2.68007215  2.67972940
#> gear        -0.27497796 -0.27514172 -0.27530005 -0.27545309 -0.27560095
#> hp          -0.04746465 -0.04745955 -0.04745451 -0.04744953 -0.04744463
#> qsec         0.15664166  0.15672110  0.15679793  0.15687220  0.15694397
#> vs           4.28877690  4.28810952  4.28745649  4.28681755  4.28619245
#> wt          -3.01727679 -3.01744553 -3.01760313 -3.01774991 -3.01788618
#>                  0.0115      0.0107      0.0100      0.0093      0.0087
#> (Intercept) 16.91491863 16.91689378 16.91881788 16.92069200 16.92251715
#> am           3.03560255  3.03583993  3.03607059  3.03629466  3.03651230
#> carb         0.83766245  0.83760895  0.83755431  0.83749859  0.83744183
#> cyl         -0.29805984 -0.29834715 -0.29862634 -0.29889760 -0.29916109
#> disp         0.02088313  0.02088401  0.02088479  0.02088548  0.02088607
#> drat         2.67939484  2.67906832  2.67874967  2.67843874  2.67813539
#> gear        -0.27574375 -0.27588163 -0.27601469 -0.27614305 -0.27626682
#> hp          -0.04743978 -0.04743501 -0.04743029 -0.04742564 -0.04742106
#> qsec         0.15701331  0.15708026  0.15714490  0.15720726  0.15726741
#> vs           4.28558094  4.28498277  4.28439771  4.28382551  4.28326595
#> wt          -3.01801224 -3.01812838 -3.01823490 -3.01833208 -3.01842019
#>                  0.0081      0.0076      0.0071      0.0066      0.0061
#> (Intercept) 16.92429436 16.92602460 16.92770887 16.92934811 16.93094328
#> am           3.03672364  3.03692882  3.03712797  3.03732124  3.03750874
#> carb         0.83738409  0.83732542  0.83726588  0.83720551  0.83714435
#> cyl         -0.29941698 -0.29966545 -0.29990664 -0.30014073 -0.30036787
#> disp         0.02088657  0.02088699  0.02088731  0.02088755  0.02088772
#> drat         2.67783945  2.67755080  2.67726927  2.67699473  2.67672703
#> gear        -0.27638611 -0.27650103 -0.27661170 -0.27671820 -0.27682065
#> hp          -0.04741653 -0.04741207 -0.04740767 -0.04740332 -0.04739904
#> qsec         0.15732540  0.15738128  0.15743510  0.15748691  0.15753677
#> vs           4.28271878  4.28218378  4.28166073  4.28114941  4.28064959
#> wt          -3.01849951 -3.01857030 -3.01863282 -3.01868732 -3.01873405
#> 
#> $iter
#>   [1]   0   7   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   7  32  28  27
#>  [37]  27  13  39   1   1  62  22   1   1   1   1   1   1   1   1   1  65   1
#>  [55]  35   1   1   1   1   1   1   1  31   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1 240   1   1   1   1  67   1   2 131   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 6.134459296 5.721017271 5.335439854 4.975849065 4.640493492 4.327739762
#>   [7] 4.036064586 3.764047341 3.510363148 3.273776421 3.053134847 2.847363777
#>  [13] 2.655460988 2.476491805 2.309584546 2.153926277 2.008758855 1.873375231
#>  [19] 1.747116011 1.629366239 1.519552408 1.417139661 1.321629191 1.232555807
#>  [25] 1.149485671 1.072014184 0.999764016 0.932383267 0.869543756 0.810939417
#>  [31] 0.756284814 0.705313749 0.657777963 0.613445930 0.572101727 0.533543985
#>  [37] 0.497584907 0.464049349 0.432773976 0.403606459 0.376404734 0.351036315
#>  [43] 0.327377642 0.305313485 0.284736378 0.265546099 0.247649180 0.230958455
#>  [49] 0.215392628 0.200875886 0.187337523 0.174711601 0.162936624 0.151955241
#>  [55] 0.141713966 0.132162919 0.123255581 0.114948567 0.107201418 0.099976402
#>  [61] 0.093238327 0.086954376 0.081093942 0.075628481 0.070531375 0.065777796
#>  [67] 0.061344593 0.057210173 0.053354399 0.049758491 0.046404935 0.043277398
#>  [73] 0.040360646 0.037640473 0.035103631 0.032737764 0.030531348 0.028473638
#>  [79] 0.026554610 0.024764918 0.023095845 0.021539263 0.020087589 0.018733752
#>  [85] 0.017471160 0.016293662 0.015195524 0.014171397 0.013216292 0.012325558
#>  [91] 0.011494857 0.010720142 0.009997640 0.009323833 0.008695438 0.008109394
#>  [97] 0.007562848 0.007053137 0.006577780 0.006134459
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 33
#> 
#> $loss
#>   [1] 960.98667 809.27999 682.35479 576.64483 489.07118 416.97720 358.07215
#>   [8] 310.38214 272.20763 242.08656 218.76230 201.15589 188.34197 179.52786
#>  [15] 174.03550 171.28573 170.72326 170.72326 170.72326 170.72326 170.72326
#>  [22] 170.72326 170.72326 170.72326 170.72326 170.72326 170.72326 170.72326
#>  [29] 170.72326 170.72326 170.72326 170.72326 169.52673 155.89623 143.93757
#>  [36] 136.58408 133.00860 132.31916 124.69310 124.69310 124.69309 107.63991
#>  [43] 107.14804 107.14804 107.14803 107.14803 107.14803 107.14803 107.14803
#>  [50] 107.14803 107.14803 107.14803 101.86398 101.86397  97.67755  97.67755
#>  [57]  97.67754  97.67754  97.67754  97.67754  97.67754  97.67754  96.96648
#>  [64]  96.96647  96.96647  96.96647  96.96647  96.96647  96.96647  96.96647
#>  [71]  96.96647  96.96647  96.96647  96.96647  96.96647  95.43419  95.43417
#>  [78]  95.43416  95.43414  95.43412  95.29118  95.29116  95.29078  95.00101
#>  [85]  95.00099  95.00097  95.00096  95.00094  95.00092  95.00091  95.00090
#>  [92]  95.00088  95.00087  95.00086  95.00085  95.00084  95.00083  95.00082
#>  [99]  95.00081  95.00080
#> 
#> $linear.predictors
#>      6.1345   5.7210   5.3354   4.9758   4.6405   4.3277   4.0361   3.7640
#> 1  20.23333 20.56171 20.86796 21.15357 21.41993 21.66834 21.90001 22.11606
#> 2  20.23333 20.80306 21.33439 21.82991 22.29203 22.72301 23.12494 23.49979
#> 3  20.23333 20.26003 20.28493 20.30815 20.32981 20.35000 20.36884 20.38640
#> 4  20.23333 19.63346 19.07402 18.55228 18.06570 17.61192 17.18872 16.79404
#> 5  20.23333 20.62344 20.98726 21.32656 21.64299 21.93809 22.21331 22.46997
#> 6  20.23333 20.65083 21.04018 21.40330 21.74194 22.05776 22.35229 22.62697
#> 7  20.23333 19.11364 18.06940 17.09555 16.18733 15.34032 14.55040 13.81371
#> 8  20.23333 19.16933 18.17704 17.25163 16.38858 15.58371 14.83307 14.13303
#> 9  20.23333 19.26216 18.35644 17.51176 16.72401 15.98935 15.30420 14.66524
#> 10 20.23333 20.93905 21.59720 22.21100 22.78343 23.31728 23.81515 24.27946
#> 11 20.23333 20.97432 21.66537 22.30985 22.91089 23.47142 23.99418 24.48170
#> 12 20.23333 20.74690 21.22586 21.67253 22.08910 22.47760 22.83991 23.17780
#> 13 20.23333 19.82839 19.45074 19.09855 18.77009 18.46377 18.17809 17.91167
#> 14 20.23333 19.89337 19.57632 19.28064 19.00489 18.74772 18.50788 18.28421
#> 15 20.23333 19.67987 19.16371 18.68234 18.23341 17.81474 17.42428 17.06014
#> 16 20.23333 20.93766 21.59451 22.20710 22.77840 23.31119 23.80808 24.27148
#> 17 20.23333 20.74597 21.22406 21.66993 22.08575 22.47354 22.83520 23.17248
#> 18 20.23333 20.86293 21.45010 21.99769 22.50838 22.98465 23.42882 23.84306
#> 19 20.23333 19.67523 19.15474 18.66933 18.21664 17.79446 17.40073 17.03353
#> 20 20.23333 20.63133 21.00251 21.34867 21.67150 21.97257 22.25335 22.51521
#> 21 20.23333 19.90729 19.60323 19.31966 19.05520 18.80856 18.57855 18.36404
#>      3.5104   3.2738   3.0531   2.8474   2.6555   2.4765    2.3096    2.1539
#> 1  22.31755 22.50546 22.68071 22.84415 22.99657 23.13872 23.271285 23.394918
#> 2  23.84936 24.17538 24.47943 24.76298 25.02743 25.27405 25.504050 25.718549
#> 3  20.40279 20.41806 20.43231 20.44560 20.45799 20.46955 20.480328 20.490380
#> 4  16.42597 16.08270 15.76256 15.46400 15.18557 14.92590 14.683725 14.457877
#> 5  22.70934 22.93258 23.14077 23.33493 23.51600 23.68487 23.842357 23.989231
#> 6  22.88314 23.12205 23.34485 23.55264 23.74642 23.92715 24.095690 24.252874
#> 7  13.12668 12.48595 11.88840 11.33112 10.81141 10.32672  9.874692  9.453134
#> 8  13.48017 12.87131 12.30349 11.77393 11.28007 10.81948 10.389946  9.989356
#> 9  14.06933 13.51359 12.99530 12.51195 12.06117 11.64077 11.248702 10.883060
#> 10 24.71248 25.11632 25.49294 25.84418 26.17174 26.47723 26.762128 27.027825
#> 11 24.93636 25.36038 25.75583 26.12462 26.46856 26.78932 27.088455 27.367433
#> 12 23.49292 23.78681 24.06088 24.31649 24.55486 24.77717 24.984503 25.177858
#> 13 17.66320 17.43148 17.21537 17.01383 16.82588 16.65059 16.487113 16.334655
#> 14 18.07561 17.88107 17.69964 17.53044 17.37265 17.22549 17.088242 16.960248
#> 15 16.72054 16.40383 16.10847 15.83301 15.57612 15.33654 15.113103 14.904729
#> 16 24.70364 25.10669 25.48256 25.83311 26.16002 26.46491 26.749246 27.014420
#> 17 23.48703 23.78038 24.05396 24.30911 24.54705 24.76896 24.975915 25.168921
#> 18 24.22937 24.58965 24.92565 25.23900 25.53124 25.80378 26.057948 26.294988
#> 19 16.69109 16.37172 16.07388 15.79611 15.53706 15.29547 15.070166 14.860044
#> 20 22.75942 22.98717 23.19957 23.39766 23.58239 23.75468 23.915352 24.065196
#> 21 18.16398 17.97741 17.80342 17.64115 17.48981 17.34868 17.217055 17.094304
#>       2.0088    1.8734    1.7471    1.6294    1.5196    1.4171    1.3216
#> 1  23.481577 23.481577 23.481577 23.481577 23.481577 23.481577 23.481577
#> 2  25.868899 25.868899 25.868899 25.868899 25.868899 25.868899 25.868899
#> 3  20.497425 20.497425 20.497425 20.497425 20.497425 20.497425 20.497425
#> 4  14.299571 14.299571 14.299571 14.299571 14.299571 14.299571 14.299571
#> 5  24.092181 24.092181 24.092181 24.092181 24.092181 24.092181 24.092181
#> 6  24.363050 24.363050 24.363050 24.363050 24.363050 24.363050 24.363050
#> 7   9.157648  9.157648  9.157648  9.157648  9.157648  9.157648  9.157648
#> 8   9.708568  9.708568  9.708568  9.708568  9.708568  9.708568  9.708568
#> 9  10.626769 10.626769 10.626769 10.626769 10.626769 10.626769 10.626769
#> 10 27.214063 27.214063 27.214063 27.214063 27.214063 27.214063 27.214063
#> 11 27.562979 27.562979 27.562979 27.562979 27.562979 27.562979 27.562979
#> 12 25.313387 25.313387 25.313387 25.313387 25.313387 25.313387 25.313387
#> 13 16.227793 16.227793 16.227793 16.227793 16.227793 16.227793 16.227793
#> 14 16.870533 16.870533 16.870533 16.870533 16.870533 16.870533 16.870533
#> 15 14.758672 14.758672 14.758672 14.758672 14.758672 14.758672 14.758672
#> 16 27.200290 27.200290 27.200290 27.200290 27.200290 27.200290 27.200290
#> 17 25.304205 25.304205 25.304205 25.304205 25.304205 25.304205 25.304205
#> 18 26.461138 26.461138 26.461138 26.461138 26.461138 26.461138 26.461138
#> 19 14.712762 14.712762 14.712762 14.712762 14.712762 14.712762 14.712762
#> 20 24.170228 24.170228 24.170228 24.170228 24.170228 24.170228 24.170228
#> 21 17.008263 17.008263 17.008263 17.008263 17.008263 17.008263 17.008263
#>       1.2326    1.1495    1.0720    0.9998    0.9324    0.8695    0.8109
#> 1  23.481577 23.481577 23.481577 23.481577 23.481577 23.481577 23.481577
#> 2  25.868899 25.868899 25.868899 25.868899 25.868899 25.868899 25.868899
#> 3  20.497425 20.497425 20.497425 20.497425 20.497425 20.497425 20.497425
#> 4  14.299571 14.299571 14.299571 14.299571 14.299571 14.299571 14.299571
#> 5  24.092181 24.092181 24.092181 24.092181 24.092181 24.092181 24.092181
#> 6  24.363050 24.363050 24.363050 24.363050 24.363050 24.363050 24.363050
#> 7   9.157648  9.157648  9.157648  9.157648  9.157648  9.157648  9.157648
#> 8   9.708568  9.708568  9.708568  9.708568  9.708568  9.708568  9.708568
#> 9  10.626769 10.626769 10.626769 10.626769 10.626769 10.626769 10.626769
#> 10 27.214063 27.214063 27.214063 27.214063 27.214063 27.214063 27.214063
#> 11 27.562979 27.562979 27.562979 27.562979 27.562979 27.562979 27.562979
#> 12 25.313387 25.313387 25.313387 25.313387 25.313387 25.313387 25.313387
#> 13 16.227793 16.227793 16.227793 16.227793 16.227793 16.227793 16.227793
#> 14 16.870533 16.870533 16.870533 16.870533 16.870533 16.870533 16.870533
#> 15 14.758672 14.758672 14.758672 14.758672 14.758672 14.758672 14.758672
#> 16 27.200290 27.200290 27.200290 27.200290 27.200290 27.200290 27.200290
#> 17 25.304205 25.304205 25.304205 25.304205 25.304205 25.304205 25.304205
#> 18 26.461138 26.461138 26.461138 26.461138 26.461138 26.461138 26.461138
#> 19 14.712762 14.712762 14.712762 14.712762 14.712762 14.712762 14.712762
#> 20 24.170228 24.170228 24.170228 24.170228 24.170228 24.170228 24.170228
#> 21 17.008263 17.008263 17.008263 17.008263 17.008263 17.008263 17.008263
#>       0.7563    0.7053    0.6578    0.6134    0.5721    0.5335    0.4976
#> 1  23.481577 23.481577 23.442446 23.294247 23.142126 23.000581 22.868314
#> 2  25.868899 25.868899 25.891567 25.987536 26.087435 26.180483 26.267355
#> 3  20.497425 20.497425 20.540233 20.232819 19.825801 19.446726 19.092785
#> 4  14.299571 14.299571 14.294866 14.211014 14.112605 14.020908 13.935326
#> 5  24.092181 24.092181 24.097927 24.053228 23.994108 23.939092 23.887687
#> 6  24.363050 24.363050 24.367781 24.455894 24.559624 24.656255 24.746461
#> 7   9.157648  9.157648  9.172222  9.192774  9.207896  9.221826  9.234957
#> 8   9.708568  9.708568  9.721076  9.756145  9.790477  9.822304  9.852141
#> 9  10.626769 10.626769 10.635834 10.767921 10.923163 11.067593 11.202572
#> 10 27.214063 27.214063 27.231688 27.401066 27.594210 27.774110 27.942069
#> 11 27.562979 27.562979 27.579296 27.819340 28.099687 28.360793 28.604581
#> 12 25.313387 25.313387 25.338139 25.368439 25.386678 25.403676 25.419538
#> 13 16.227793 16.227793 16.263025 16.055599 15.776577 15.516652 15.274009
#> 14 16.870533 16.870533 16.903356 16.910991 16.896221 16.882371 16.869516
#> 15 14.758672 14.758672 14.752245 14.977142 15.256848 15.517250 15.760468
#> 16 27.200290 27.200290 27.217966 27.388106 27.582143 27.762874 27.931609
#> 17 25.304205 25.304205 25.305408 25.669189 26.111959 26.524309 26.909337
#> 18 26.461138 26.461138 26.458003 26.333717 26.183858 26.044391 25.914087
#> 19 14.712762 14.712762 14.706507 15.248801 15.915812 16.536861 17.116869
#> 20 24.170228 24.170228 24.081347 23.442948 22.729316 22.065002 21.444471
#> 21 17.008263 17.008263 16.899068 16.333082 15.723457 15.155944 14.625849
#>       0.4640    0.4328    0.4036    0.3764   0.3510   0.3274   0.3053   0.2847
#> 1  22.773434 23.119390 23.119418 23.119437 22.01953 21.96850 21.96850 21.96852
#> 2  26.325450 27.811584 27.811230 27.810993 28.04812 27.96219 27.96256 27.96283
#> 3  19.034233 20.603210 20.603236 20.603253 20.34529 20.04190 20.04209 20.04223
#> 4  13.901417 13.365850 13.366027 13.366145 13.11340 13.11453 13.11437 13.11426
#> 5  23.878457 21.971709 21.972128 21.972408 22.78119 22.89655 22.89610 22.89576
#> 6  24.781957 22.152630 22.153066 22.153357 23.27465 23.50159 23.50104 23.50063
#> 7   9.258805  9.931408  9.931273  9.931183 10.19573 10.14593 10.14613 10.14624
#> 8   9.878677 10.299384 10.299283 10.299215 10.57148 10.54737 10.54751 10.54759
#> 9  11.258740 10.912677 10.912631 10.912601 11.37661 11.45266 11.45268 11.45266
#> 10 28.019841 28.710058 28.709786 28.709604 29.05882 29.06552 29.06574 29.06590
#> 11 28.705287 28.943110 28.942859 28.942691 29.44779 29.51917 29.51932 29.51943
#> 12 25.457298 23.819932 23.820250 23.820462 24.14778 24.23499 24.23460 24.23432
#> 13 15.241766 16.718853 16.718796 16.718757 14.85714 14.57324 14.57343 14.57359
#> 14 16.908835 17.148158 17.148140 17.148127 15.78219 15.68427 15.68428 15.68431
#> 15 15.831254 13.672496 13.672701 13.672838 14.15522 14.41134 14.41098 14.41070
#> 16 28.009702 28.700859 28.700586 28.700403 29.05219 29.05913 29.05935 29.05952
#> 17 27.036068 26.401865 26.401652 26.401510 25.25929 25.42092 25.42090 25.42092
#> 18 25.867374 27.174614 27.174472 27.174377 27.44417 27.31897 27.31921 27.31939
#> 19 17.297467 17.262441 17.261938 17.261601 18.30266 18.56827 18.56856 18.56875
#> 20 21.124905 21.514273 21.514693 21.514975 20.38349 20.20198 20.20178 20.20166
#> 21 14.309034 14.665499 14.665836 14.666062 15.28326 15.21098 15.21089 15.21080
#>      0.2655   0.2476   0.2310   0.2154   0.2009   0.1873   0.1747   0.1629
#> 1  21.96855 21.96859 21.96864 21.96867 21.96871 21.96873 21.96876 23.09712
#> 2  27.96302 27.96316 27.96326 27.96332 27.96336 27.96338 27.96339 27.53502
#> 3  20.04233 20.04240 20.04245 20.04249 20.04251 20.04252 20.04253 19.69296
#> 4  13.11417 13.11411 13.11407 13.11404 13.11402 13.11401 13.11400 12.18083
#> 5  22.89551 22.89533 22.89521 22.89512 22.89507 22.89504 22.89502 24.10425
#> 6  23.50032 23.50010 23.49994 23.49984 23.49977 23.49973 23.49970 23.89694
#> 7  10.14631 10.14634 10.14635 10.14635 10.14634 10.14632 10.14631 11.33530
#> 8  10.54764 10.54765 10.54765 10.54764 10.54762 10.54761 10.54759 11.37979
#> 9  11.45264 11.45260 11.45257 11.45254 11.45251 11.45248 11.45246 11.79718
#> 10 29.06602 29.06611 29.06617 29.06621 29.06623 29.06625 29.06626 29.16388
#> 11 29.51951 29.51957 29.51961 29.51964 29.51965 29.51966 29.51967 29.63904
#> 12 24.23413 24.23399 24.23390 24.23385 24.23381 24.23379 24.23379 23.47936
#> 13 14.57373 14.57384 14.57393 14.57400 14.57405 14.57409 14.57412 13.95251
#> 14 15.68434 15.68438 15.68441 15.68445 15.68447 15.68449 15.68451 15.12746
#> 15 14.41048 14.41032 14.41021 14.41013 14.41007 14.41004 14.41001 13.65703
#> 16 29.05964 29.05972 29.05978 29.05982 29.05985 29.05986 29.05987 29.16075
#> 17 25.42095 25.42099 25.42103 25.42107 25.42111 25.42114 25.42116 25.41365
#> 18 27.31952 27.31961 27.31968 27.31972 27.31975 27.31976 27.31977 26.92781
#> 19 18.56886 18.56892 18.56895 18.56895 18.56894 18.56893 18.56891 17.86275
#> 20 20.20160 20.20157 20.20157 20.20158 20.20160 20.20162 20.20164 20.79009
#> 21 15.21073 15.21067 15.21063 15.21059 15.21056 15.21054 15.21052 14.70628
#>      0.1520   0.1417   0.1322   0.1233   0.1149   0.1072   0.1000   0.0932
#> 1  23.09707 22.65133 22.65178 22.65208 22.65228 22.65241 22.65249 22.65254
#> 2  27.53526 27.44318 27.44312 27.44311 27.44314 27.44317 27.44321 27.44324
#> 3  19.69309 19.69996 19.70004 19.70010 19.70014 19.70017 19.70019 19.70021
#> 4  12.18067 12.86653 12.86649 12.86639 12.86629 12.86619 12.86611 12.86605
#> 5  24.10402 23.70043 23.70025 23.70017 23.70013 23.70012 23.70011 23.70012
#> 6  23.89659 23.36878 23.36837 23.36812 23.36798 23.36790 23.36786 23.36783
#> 7  11.33564 11.21869 11.21881 11.21892 11.21900 11.21907 11.21911 11.21915
#> 8  11.38007 10.88548 10.88552 10.88558 10.88564 10.88569 10.88573 10.88577
#> 9  11.79732 11.13384 11.13372 11.13367 11.13366 11.13367 11.13368 11.13370
#> 10 29.16403 28.98218 28.98212 28.98213 28.98216 28.98220 28.98224 28.98228
#> 11 29.63914 29.82982 29.82972 29.82969 29.82969 29.82970 29.82972 29.82973
#> 12 23.47898 23.60300 23.60273 23.60254 23.60241 23.60231 23.60225 23.60220
#> 13 13.95245 14.57428 14.57483 14.57513 14.57528 14.57535 14.57537 14.57537
#> 14 15.12728 15.68445 15.68482 15.68501 15.68509 15.68512 15.68512 15.68510
#> 15 13.65674 13.86587 13.86555 13.86531 13.86513 13.86500 13.86491 13.86484
#> 16 29.16090 29.31462 29.31458 29.31458 29.31461 29.31464 29.31466 29.31469
#> 17 25.41341 25.49775 25.49808 25.49828 25.49840 25.49847 25.49850 25.49852
#> 18 26.92799 27.65961 27.65956 27.65952 27.65949 27.65946 27.65944 27.65943
#> 19 17.86304 18.56093 18.56060 18.56038 18.56024 18.56015 18.56009 18.56006
#> 20 20.78993 20.14475 20.14509 20.14528 20.14540 20.14546 20.14550 20.14552
#> 21 14.70639 14.21453 14.21423 14.21400 14.21385 14.21375 14.21368 14.21365
#>      0.0870   0.0811   0.0756   0.0705   0.0658   0.0613   0.0572   0.0534
#> 1  22.65257 22.87235 22.87220 22.87209 22.87200 22.87193 22.87188 22.87185
#> 2  27.44326 27.23098 27.23085 27.23078 27.23077 27.23078 27.23081 27.23085
#> 3  19.70021 19.62569 19.62577 19.62583 19.62587 19.62590 19.62591 19.62592
#> 4  12.86600 12.83297 12.83340 12.83370 12.83390 12.83403 12.83410 12.83414
#> 5  23.70012 24.02746 24.02714 24.02686 24.02662 24.02642 24.02627 24.02614
#> 6  23.36782 23.58754 23.58725 23.58701 23.58681 23.58664 23.58651 23.58642
#> 7  11.21917 11.21421 11.21408 11.21399 11.21393 11.21390 11.21389 11.21389
#> 8  10.88580 10.83655 10.83637 10.83625 10.83618 10.83614 10.83613 10.83614
#> 9  11.13372 11.04156 11.04138 11.04126 11.04120 11.04117 11.04117 11.04118
#> 10 28.98231 28.87498 28.87474 28.87459 28.87451 28.87447 28.87446 28.87447
#> 11 29.82975 29.75123 29.75106 29.75096 29.75090 29.75088 29.75087 29.75088
#> 12 23.60216 23.63419 23.63429 23.63434 23.63437 23.63437 23.63435 23.63433
#> 13 14.57536 14.42641 14.42693 14.42732 14.42761 14.42782 14.42796 14.42806
#> 14 15.68509 15.54830 15.54872 15.54905 15.54929 15.54947 15.54960 15.54969
#> 15 13.86479 13.82393 13.82412 13.82425 13.82434 13.82440 13.82444 13.82446
#> 16 29.31471 29.22401 29.22385 29.22376 29.22371 29.22369 29.22369 29.22370
#> 17 25.49853 25.39962 25.39968 25.39977 25.39987 25.39996 25.40004 25.40012
#> 18 27.65943 27.64709 27.64720 27.64727 27.64731 27.64733 27.64734 27.64733
#> 19 18.56005 18.23420 18.23429 18.23440 18.23452 18.23463 18.23473 18.23482
#> 20 20.14552 20.58194 20.58191 20.58182 20.58171 20.58159 20.58147 20.58137
#> 21 14.21363 14.48478 14.48477 14.48470 14.48459 14.48448 14.48437 14.48427
#>      0.0498   0.0464   0.0433   0.0404   0.0376   0.0351   0.0327   0.0305
#> 1  22.87182 22.87181 22.87180 22.87179 22.87179 22.87179 22.94677 22.94679
#> 2  27.23089 27.23093 27.23097 27.23100 27.23102 27.23104 26.85225 26.85213
#> 3  19.62592 19.62592 19.62592 19.62592 19.62592 19.62591 19.39947 19.39941
#> 4  12.83414 12.83414 12.83413 12.83411 12.83409 12.83407 13.01937 13.01945
#> 5  24.02605 24.02598 24.02593 24.02590 24.02587 24.02586 24.40413 24.40423
#> 6  23.58634 23.58629 23.58625 23.58622 23.58620 23.58619 23.46965 23.46959
#> 7  11.21389 11.21390 11.21391 11.21392 11.21392 11.21393 11.75946 11.75963
#> 8  10.83615 10.83616 10.83618 10.83620 10.83621 10.83622 10.86227 10.86227
#> 9  11.04120 11.04122 11.04124 11.04126 11.04128 11.04129 10.68704 10.68692
#> 10 28.87449 28.87451 28.87454 28.87456 28.87458 28.87460 28.69018 28.69011
#> 11 29.75089 29.75091 29.75092 29.75094 29.75096 29.75097 29.92860 29.92864
#> 12 23.63431 23.63429 23.63427 23.63425 23.63423 23.63422 23.36954 23.36945
#> 13 14.42813 14.42817 14.42819 14.42820 14.42820 14.42819 14.38423 14.38424
#> 14 15.54975 15.54979 15.54981 15.54982 15.54982 15.54982 15.51943 15.51943
#> 15 13.82446 13.82446 13.82446 13.82445 13.82444 13.82444 13.72509 13.72505
#> 16 29.22372 29.22374 29.22376 29.22377 29.22379 29.22380 29.34581 29.34584
#> 17 25.40018 25.40023 25.40027 25.40030 25.40033 25.40035 25.18052 25.18044
#> 18 27.64732 27.64731 27.64730 27.64729 27.64727 27.64727 28.14295 28.14312
#> 19 18.23490 18.23496 18.23501 18.23505 18.23508 18.23510 18.56994 18.57004
#> 20 20.58128 20.58120 20.58114 20.58109 20.58105 20.58102 20.41252 20.41249
#> 21 14.48418 14.48410 14.48404 14.48398 14.48394 14.48391 14.23079 14.23073
#>      0.0285   0.0266   0.0248   0.0231   0.0215   0.0201   0.0187   0.0175
#> 1  22.94680 22.94682 22.94683 22.97211 22.97207 22.97220 22.97856 22.97848
#> 2  26.85201 26.85189 26.85177 26.78353 26.78344 26.78347 26.85705 26.85718
#> 3  19.39936 19.39930 19.39925 19.40180 19.40185 19.40205 19.26823 19.26815
#> 4  13.01953 13.01961 13.01969 13.05398 13.05415 13.05461 13.26157 13.26175
#> 5  24.40433 24.40442 24.40452 24.25526 24.25479 24.25359 24.10454 24.10425
#> 6  23.46953 23.46947 23.46942 23.68876 23.68902 23.68919 23.60600 23.60604
#> 7  11.75980 11.75996 11.76012 11.83357 11.83372 11.83402 12.01396 12.01409
#> 8  10.86226 10.86226 10.86226 10.85409 10.85406 10.85403 10.92648 10.92656
#> 9  10.68679 10.68667 10.68655 10.60696 10.60677 10.60647 10.58947 10.58948
#> 10 28.69004 28.68997 28.68990 28.67910 28.67906 28.67919 28.64071 28.64066
#> 11 29.92868 29.92873 29.92877 30.03856 30.03878 30.03939 30.07346 30.07345
#> 12 23.36937 23.36929 23.36921 23.34519 23.34510 23.34513 23.53994 23.54015
#> 13 14.38426 14.38427 14.38428 14.40837 14.40852 14.40889 14.32415 14.32412
#> 14 15.51944 15.51944 15.51944 15.56801 15.56817 15.56857 15.40159 15.40143
#> 15 13.72502 13.72499 13.72495 13.58589 13.58563 13.58530 13.56558 13.56545
#> 16 29.34587 29.34590 29.34593 29.32871 29.32868 29.32885 29.35726 29.35721
#> 17 25.18035 25.18026 25.18018 25.17230 25.17224 25.17185 25.28164 25.28191
#> 18 28.14330 28.14347 28.14363 28.08313 28.08310 28.08282 28.15416 28.15419
#> 19 18.57014 18.57024 18.57034 18.58851 18.58869 18.58871 18.36197 18.36174
#> 20 20.41245 20.41241 20.41237 20.36386 20.36368 20.36313 20.30894 20.30890
#> 21 14.23068 14.23062 14.23057 14.28830 14.28846 14.28854 14.28473 14.28481
#>      0.0163   0.0152   0.0142   0.0132   0.0123   0.0115   0.0107   0.0100
#> 1  22.97839 22.97831 22.97822 22.97814 22.97806 22.97798 22.97790 22.97783
#> 2  26.85731 26.85743 26.85756 26.85768 26.85781 26.85793 26.85805 26.85816
#> 3  19.26807 19.26799 19.26792 19.26785 19.26778 19.26771 19.26765 19.26759
#> 4  13.26192 13.26208 13.26224 13.26239 13.26254 13.26268 13.26282 13.26295
#> 5  24.10397 24.10370 24.10343 24.10317 24.10292 24.10267 24.10243 24.10220
#> 6  23.60608 23.60612 23.60616 23.60620 23.60623 23.60627 23.60630 23.60634
#> 7  12.01420 12.01431 12.01442 12.01452 12.01461 12.01470 12.01478 12.01486
#> 8  10.92664 10.92672 10.92679 10.92687 10.92694 10.92701 10.92707 10.92714
#> 9  10.58950 10.58953 10.58955 10.58957 10.58960 10.58962 10.58965 10.58968
#> 10 28.64060 28.64055 28.64051 28.64046 28.64042 28.64038 28.64034 28.64030
#> 11 30.07343 30.07341 30.07338 30.07336 30.07334 30.07331 30.07329 30.07326
#> 12 23.54035 23.54055 23.54074 23.54093 23.54111 23.54129 23.54147 23.54164
#> 13 14.32409 14.32406 14.32403 14.32400 14.32397 14.32394 14.32391 14.32389
#> 14 15.40128 15.40113 15.40099 15.40085 15.40072 15.40059 15.40046 15.40034
#> 15 13.56532 13.56520 13.56509 13.56497 13.56486 13.56476 13.56465 13.56455
#> 16 29.35716 29.35711 29.35706 29.35701 29.35697 29.35692 29.35687 29.35683
#> 17 25.28217 25.28243 25.28268 25.28293 25.28317 25.28340 25.28363 25.28386
#> 18 28.15423 28.15426 28.15428 28.15431 28.15433 28.15435 28.15437 28.15438
#> 19 18.36152 18.36131 18.36109 18.36089 18.36069 18.36049 18.36030 18.36011
#> 20 20.30887 20.30884 20.30881 20.30878 20.30876 20.30873 20.30871 20.30869
#> 21 14.28489 14.28497 14.28504 14.28512 14.28519 14.28526 14.28533 14.28540
#>      0.0093   0.0087   0.0081   0.0076   0.0071   0.0066   0.0061
#> 1  22.97775 22.97768 22.97761 22.97754 22.97747 22.97740 22.97733
#> 2  26.85828 26.85840 26.85851 26.85862 26.85873 26.85884 26.85895
#> 3  19.26753 19.26748 19.26743 19.26738 19.26733 19.26729 19.26724
#> 4  13.26307 13.26319 13.26331 13.26342 13.26353 13.26363 13.26372
#> 5  24.10197 24.10174 24.10153 24.10131 24.10111 24.10091 24.10071
#> 6  23.60637 23.60640 23.60644 23.60647 23.60650 23.60653 23.60656
#> 7  12.01494 12.01500 12.01507 12.01513 12.01518 12.01523 12.01528
#> 8  10.92720 10.92726 10.92732 10.92738 10.92744 10.92749 10.92754
#> 9  10.58971 10.58974 10.58978 10.58981 10.58984 10.58988 10.58992
#> 10 28.64027 28.64023 28.64020 28.64017 28.64014 28.64012 28.64009
#> 11 30.07324 30.07321 30.07318 30.07315 30.07312 30.07309 30.07306
#> 12 23.54181 23.54197 23.54213 23.54229 23.54244 23.54259 23.54273
#> 13 14.32386 14.32384 14.32382 14.32379 14.32377 14.32375 14.32373
#> 14 15.40022 15.40010 15.39998 15.39987 15.39977 15.39966 15.39956
#> 15 13.56446 13.56436 13.56427 13.56419 13.56410 13.56402 13.56395
#> 16 29.35678 29.35674 29.35670 29.35665 29.35661 29.35657 29.35653
#> 17 25.28408 25.28430 25.28451 25.28472 25.28492 25.28512 25.28532
#> 18 28.15440 28.15441 28.15442 28.15442 28.15443 28.15443 28.15443
#> 19 18.35992 18.35974 18.35957 18.35940 18.35923 18.35907 18.35891
#> 20 20.30867 20.30865 20.30863 20.30861 20.30860 20.30858 20.30857
#> 21 14.28547 14.28554 14.28560 14.28567 14.28573 14.28579 14.28585
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 22.8 18.1 14.3 24.4 22.8 10.4 10.4 14.7 32.4 33.9 21.5 15.5 15.2 13.3
#> [16] 27.3 26.0 30.4 15.8 19.7 15.0
#> 
#> $X
#>               am       carb       cyl        disp        drat       gear
#>  [1,]  1.0488088  0.6234292  0.000000 -0.52950777  0.59920739  0.2399041
#>  [2,]  1.0488088 -1.0130725 -1.080123 -0.91867353  0.49623395  0.2399041
#>  [3,] -0.9534626 -1.0130725  0.000000 -0.04305057 -1.74858721 -1.0195922
#>  [4,] -0.9534626  0.6234292  1.080123  0.96728361 -0.82182618 -1.0195922
#>  [5,] -0.9534626 -0.4675719 -1.080123 -0.62904440  0.16671891  0.2399041
#>  [6,] -0.9534626 -0.4675719 -1.080123 -0.67319974  0.64039677  0.2399041
#>  [7,] -0.9534626  0.6234292  1.080123  1.80548678 -1.39847748 -1.0195922
#>  [8,] -0.9534626  0.6234292  1.080123  1.71567930 -1.25431466 -1.0195922
#>  [9,] -0.9534626  0.6234292  1.080123  1.56600016 -0.78063680 -1.0195922
#> [10,]  1.0488088 -1.0130725 -1.080123 -1.13795347  0.96991180  0.2399041
#> [11,]  1.0488088 -1.0130725 -1.080123 -1.19483154  1.25823746  0.2399041
#> [12,] -0.9534626 -1.0130725 -1.080123 -0.82811765  0.18731360 -1.0195922
#> [13,] -0.9534626 -0.4675719  1.080123  0.65295742 -1.74858721 -1.0195922
#> [14,] -0.9534626 -0.4675719  1.080123  0.54818202 -0.94539432 -1.0195922
#> [15,] -0.9534626  0.6234292  1.080123  0.89244404  0.24909767 -1.0195922
#> [16,]  1.0488088 -1.0130725 -1.080123 -1.13570828  0.96991180  0.2399041
#> [17,]  1.0488088 -0.4675719 -1.080123 -0.82662086  1.69072594  1.4994004
#> [18,]  1.0488088 -0.4675719 -1.080123 -1.01521657  0.33147643  1.4994004
#> [19,]  1.0488088  0.6234292  1.080123  0.89992800  1.25823746  1.4994004
#> [20,]  1.0488088  1.7144303  0.000000 -0.64176712  0.02255609  1.4994004
#> [21,]  1.0488088  2.8054314  1.080123  0.52573015 -0.14220143  1.4994004
#>                hp         qsec         vs            wt
#>  [1,] -0.53267870 -0.377236152 -0.8660254 -0.2898933543
#>  [2,] -0.75201699  0.388472974  1.1547005 -0.8016635821
#>  [3,] -0.59718996  1.163813661  1.1547005  0.2495401291
#>  [4,]  1.20912535 -0.945498271 -0.8660254  0.3509720661
#>  [5,] -1.15198681  1.057866487  1.1547005  0.0005708291
#>  [6,] -0.72621248  2.454442881  1.1547005 -0.0363135117
#>  [7,]  0.69303526  0.085078792 -0.8660254  1.9001143773
#>  [8,]  0.82205779  0.008026301 -0.8660254  2.0605612596
#>  [9,]  1.01559157 -0.184604926 -0.8660254  1.9877146866
#> [10,] -1.10037780  0.802630111  1.1547005 -0.9123166043
#> [11,] -1.11328005  1.009708680  1.1547005 -1.2488862136
#> [12,] -0.70040798  1.062682267  1.1547005 -0.6679578469
#> [13,] -0.01658861 -0.449472862 -0.8660254  0.3048666402
#> [14,] -0.01658861 -0.242394294 -0.8660254  0.2264874161
#> [15,]  1.20912535 -1.152576840 -0.8660254  0.5999413662
#> [16,] -1.10037780  0.528130613  1.1547005 -1.1566753617
#> [17,] -0.77782149 -0.531341134 -0.8660254 -0.9676431154
#> [18,] -0.49397194 -0.435025520  1.1547005 -1.5458051566
#> [19,]  1.45426815 -1.590812881 -0.8660254 -0.0178713413
#> [20,]  0.30596770 -1.109234814 -0.8660254 -0.3867147487
#> [21,]  2.37032805 -1.542655074 -0.8660254  0.3509720661
#> attr(,"center")
#>  [1]   0.4761905   2.8571429   6.0000000 230.7523810   3.6090476   3.8095238
#>  [7] 151.2857143  17.8033333   0.4285714   3.1893810
#> attr(,"scale")
#>  [1]   0.4994328   1.8331787   1.8516402 133.6191554   0.4855621   0.7939682
#>  [7]  77.5058479   2.0765065   0.4948717   1.0844711
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 31
#> 
#> $lambda.min
#> [1] 0.7562848
#> 
#> $null.dev
#> [1] 45.76127
#> 
#> $Bias
#> [1] 4.985313
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 14.32692