Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 25.66363 24.54070 23.43219 22.48192 21.61555 20.17431 18.98852 18.02034
#>   [9] 17.23717 16.61095 16.11752 15.73615 15.44900 15.24082 15.09850 15.00270
#>  [17] 14.62564 14.27992 13.97996 13.59972 13.35930 13.24911 13.05261 12.54150
#>  [25] 12.17315 11.91636 11.75082 11.76829 11.83769 12.81624 12.78781 16.92114
#>  [33] 16.69767 16.47719 16.95997 17.12784 15.56445 15.23160 14.65939 14.68229
#>  [41] 14.72203 14.77299 14.34755 15.95172 16.00913 16.00946 16.00977 17.85249
#>  [49] 20.70862 22.06683 22.06827 21.93416 23.46673 23.44007 24.58367 24.64145
#>  [57] 24.61575 25.15557 25.77409 25.87450 26.49126 26.49124 26.46162 25.98375
#>  [65] 25.98280 25.98195 25.98119 25.98050 25.97989 25.90648 25.84997 25.85024
#>  [73] 25.84501 25.84461 27.59690 27.59699 27.59708 27.71456 27.71464 28.05832
#>  [81] 28.06026 28.06209 28.06382 28.42291 29.44221 29.42764 29.42980 29.43190
#>  [89] 29.43393 29.83839 29.84047 29.84250 29.83880 29.84036 29.84191 29.84345
#>  [97] 29.84496 29.81722 29.83953 29.84134
#> 
#> $cvse
#>   [1] 11.609185 11.642727 11.600578 11.532124 11.341205 10.486655  9.711473
#>   [8]  9.008516  8.372330  7.798769  7.284668  6.827559  6.425403  6.076352
#>  [15]  5.778518  5.530350  5.272552  5.038617  4.847044  4.684570  4.600788
#>  [22]  4.611263  4.610567  4.504599  4.427499  4.382185  4.353936  4.324811
#>  [29]  4.296275  4.323649  4.238803  5.675725  5.642259  5.564853  5.444620
#>  [36]  5.332280  4.602139  4.559795  4.467611  4.454652  4.453158  4.474997
#>  [43]  4.467068  4.598578  4.608917  4.608903  4.608903  4.743097  5.568850
#>  [50]  5.969976  5.970296  5.954716  6.736924  6.688317  6.825040  6.822882
#>  [57]  6.826889  6.969653  7.257368  7.251464  7.353648  7.353597  7.356464
#>  [64]  7.306949  7.306761  7.306595  7.306450  7.306323  7.306213  7.267101
#>  [71]  7.271541  7.271516  7.264652  7.264269  7.298072  7.297821  7.297588
#>  [78]  7.334060  7.333847  7.580177  7.580931  7.581638  7.582302  7.627971
#>  [85]  8.289872  8.293783  8.294789  8.295766  8.296714  8.364575  8.365512
#>  [92]  8.366425  8.369107  8.369797  8.370480  8.371154  8.371811  8.361389
#>  [99]  8.374696  8.375465
#> 
#> $fold
#>  [1]  8  5  1  4  1  7  9  2  3  5  4  9  2  1  8 10  7  6  6  3 10
#> 
#> $lambda
#>   [1] 3.937126815 3.671777645 3.424312121 3.193524945 2.978292052 2.777565135
#>   [7] 2.590366541 2.415784506 2.252968716 2.101126165 1.959517294 1.827452386
#>  [13] 1.704288212 1.589424890 1.482302972 1.382400713 1.289231532 1.202341641
#>  [19] 1.121307839 1.045735443 0.975256374 0.909527359 0.848228259 0.791060513
#>  [25] 0.737745682 0.688024092 0.641653571 0.598408268 0.558077554 0.520464994
#>  [31] 0.485387396 0.452673910 0.422165203 0.393712682 0.367177765 0.342431212
#>  [37] 0.319352494 0.297829205 0.277756514 0.259036654 0.241578451 0.225296872
#>  [43] 0.210112616 0.195951729 0.182745239 0.170428821 0.158942489 0.148230297
#>  [49] 0.138240071 0.128923153 0.120234164 0.112130784 0.104573544 0.097525637
#>  [55] 0.090952736 0.084822826 0.079106051 0.073774568 0.068802409 0.064165357
#>  [61] 0.059840827 0.055807755 0.052046499 0.048538740 0.045267391 0.042216520
#>  [67] 0.039371268 0.036717776 0.034243121 0.031935249 0.029782921 0.027775651
#>  [73] 0.025903665 0.024157845 0.022529687 0.021011262 0.019595173 0.018274524
#>  [79] 0.017042882 0.015894249 0.014823030 0.013824007 0.012892315 0.012023416
#>  [85] 0.011213078 0.010457354 0.009752564 0.009095274 0.008482283 0.007910605
#>  [91] 0.007377457 0.006880241 0.006416536 0.005984083 0.005580776 0.005204650
#>  [97] 0.004853874 0.004526739 0.004221652 0.003937127
#> 
#> $fit
#> $beta
#>               3.9371     3.6718     3.4243    3.1935    2.9783    2.7776
#> (Intercept) 19.65238 21.2409134 22.7223840 24.104009 25.392516 26.594183
#> am           0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> wt           0.00000 -0.4742899 -0.9166143 -1.329127 -1.713839 -2.072622
#>                2.5904    2.4158    2.2530    2.1011    1.9595   1.8275
#> (Intercept) 27.714861 28.760009 29.734718 30.643735 31.491487 32.28210
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> wt          -2.407224 -2.719275 -3.010295 -3.281701 -3.534815 -3.77087
#>                1.7043    1.5894    1.4823    1.3824    1.2892    1.2023
#> (Intercept) 33.019434 33.707072 34.348365 34.946438 35.365648 35.365648
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -3.991016 -4.196325 -4.387797 -4.566364 -4.691528 -4.691528
#>                 1.1213     1.0457     0.9753     0.9095     0.8482     0.7911
#> (Intercept) 34.0391985 32.6456650 31.3460440 30.1340131 29.0036691 27.9495064
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0694463  0.1424048  0.2104465  0.2739024  0.3330816  0.3882723
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -4.6710522 -4.6495411 -4.6294796 -4.6107703 -4.5933218 -4.5770494
#>                 0.7377     0.6880     0.6417    0.5984     0.5581     0.5205
#> (Intercept) 26.9663908 26.0495339 25.1944701 24.397035 22.4638441 20.1545359
#> am           0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.000000  0.3437064  0.8102968
#> gear         0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> qsec         0.4397433  0.4877454  0.5325123  0.574262  0.5878906  0.5898974
#> vs           0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> wt          -4.5618736 -4.5477206 -4.5345215 -4.522212 -4.3799453 -4.1916778
#>                 0.4854    0.4527    0.4222    0.3937    0.3672    0.3424
#> (Intercept) 18.0029435 14.743402 14.741048 14.739610 14.738731 14.738194
#> am           0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.0000000 -1.117757 -1.117846 -1.117901 -1.117934 -1.117954
#> cyl          0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         1.2448782  4.481403  4.481907  4.482214  4.482402  4.482517
#> gear         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.5918035  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -4.0163149 -2.322412 -2.322165 -2.322013 -2.321921 -2.321865
#>                0.3194    0.2978    0.2778    0.2590      0.2416     0.2253
#> (Intercept) 14.737866 14.737665 14.737543 14.737468 14.74035202  9.0076308
#> am           0.000000  0.000000  0.000000  0.000000  0.03083217  2.8623940
#> carb        -1.117967 -1.117974 -1.117979 -1.117982 -1.12216143 -1.0147374
#> cyl          0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> drat         4.482587  4.482630  4.482656  4.482672  4.47390163  2.5079731
#> gear         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.00000000  0.5823497
#> vs           0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> wt          -2.321830 -2.321809 -2.321796 -2.321788 -2.31261579 -2.0143255
#>                 0.2101    0.1960     0.1827     0.1704     0.1589     0.1482
#> (Intercept)  9.0067444  9.006073  7.2328827  7.2324289  7.2320565  7.2317519
#> am           2.8631046  2.863638  2.6639800  2.6645732  2.6650516  2.6654366
#> carb        -1.0147041 -1.014680 -1.4043092 -1.4043230 -1.4043326 -1.4043393
#> cyl          0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         2.5074753  2.507104  1.1614330  1.1607246  1.1601619  1.1597156
#> gear         0.0000000  0.000000  1.7632494  1.7634441  1.7635947  1.7637112
#> hp           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.5824678  0.582556  0.5781025  0.5782104  0.5782971  0.5783667
#> vs           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.0142639 -2.014217 -1.5883155 -1.5882655 -1.5882255 -1.5881935
#>                 0.1382    0.1289     0.1202     0.1121     0.1046     0.0975
#> (Intercept)  7.2315036  7.231302  7.2311380  7.2310057  7.2308988  7.2308128
#> am           2.6657457  2.665993  2.6661916  2.6663498  2.6664758  2.6665761
#> carb        -1.4043439 -1.404347 -1.4043490 -1.4043503 -1.4043510 -1.4043514
#> cyl          0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         1.1593623  1.159083  1.1588623  1.1586883  1.1585514  1.1584437
#> gear         1.7638011  1.763870  1.7639239  1.7639650  1.7639965  1.7640208
#> hp           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.5784224  0.578467  0.5785026  0.5785309  0.5785534  0.5785713
#> vs           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -1.5881679 -1.588148 -1.5881313 -1.5881185 -1.5881082 -1.5881001
#>                 0.0910     0.0848     0.0791     0.0738     0.0688     0.0642
#> (Intercept)  7.2307438 -6.0261828 -6.0221090 -6.0182141 -6.0145116 -6.0110105
#> am           2.6666558  3.5295230  3.5299245  3.5302417  3.5304865  3.5306694
#> carb        -1.4043516 -1.6925318 -1.6925824 -1.6926170 -1.6926384 -1.6926487
#> cyl          0.0000000  0.9294320  0.9292918  0.9291456  0.9289963  0.9288463
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         1.1583591  1.5348484  1.5340540  1.5333757  1.5327996  1.5323135
#> gear         1.7640394  2.6771912  2.6773296  2.6774160  2.6774592  2.6774671
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.5785855  0.7341389  0.7340551  0.7339752  0.7338995  0.7338280
#> vs           0.0000000  1.3453124  1.3456243  1.3458654  1.3460464  1.3461765
#> wt          -1.5880937 -1.5738638 -1.5737046 -1.5735664 -1.5734466 -1.5733433
#>                 0.0598     0.0558     0.0520     0.0485     0.0453     0.0422
#> (Intercept) -6.0077153 -6.0046276 -6.0017461 -5.9990671 -5.9965852 -5.9942939
#> am           3.5307999  3.5308861  3.5309354  3.5309542  3.5309479  3.5309215
#> carb        -1.6926501 -1.6926443 -1.6926327 -1.6926166 -1.6925971 -1.6925751
#> cyl          0.9286976  0.9285517  0.9284099  0.9282733  0.9281425  0.9280180
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         1.5319063  1.5315679  1.5312894  1.5310629  1.5308812  1.5307380
#> gear         2.6774465  2.6774031  2.6773419  2.6772672  2.6771825  2.6770907
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.7337608  0.7336979  0.7336393  0.7335847  0.7335342  0.7334876
#> vs           1.3462636  1.3463150  1.3463368  1.3463343  1.3463122  1.3462746
#> wt          -1.5732546 -1.5731787 -1.5731141 -1.5730594 -1.5730135 -1.5729752
#>                   0.0394       0.0367       0.0342       0.0319       0.0298
#> (Intercept) -6.464944236 -6.461991292 -6.459290971 -6.456823989 -6.454572261
#> am           3.375285494  3.374885986  3.374514041  3.374168070  3.373846547
#> carb        -1.456901631 -1.456729780 -1.456570981 -1.456424287 -1.456288811
#> cyl          0.807193650  0.806947578  0.806721136  0.806512904  0.806321549
#> disp         0.006756201  0.006757365  0.006758442  0.006759439  0.006760365
#> drat         1.746250745  1.746498600  1.746733169  1.746954831  1.747163990
#> gear         2.535647466  2.535210558  2.534806615  2.534433422  2.534088881
#> hp           0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> qsec         0.798382025  0.798351450  0.798323579  0.798298226  0.798275213
#> vs           1.482179504  1.481826942  1.481499506  1.481195729  1.480914201
#> wt          -2.299484691 -2.299682406 -2.299865732 -2.300035805 -2.300193673
#>                   0.0278       0.0259       0.0242      0.0225       0.0210
#> (Intercept) -6.452518852 -6.450647929 -6.448944713 -6.44739543 -6.445987248
#> am           3.373548005  3.373271039  3.373014306  3.37277652  3.372556475
#> carb        -1.456163726 -1.456048256 -1.455941680 -1.45584332 -1.455752562
#> cyl          0.806145817  0.805984534  0.805836598  0.80570098  0.805576717
#> disp         0.006761224  0.006762023  0.006762766  0.00676346  0.006764107
#> drat         1.747361071  1.747546511  1.747720759  1.74788427  1.748037508
#> gear         2.533771011  2.533477941  2.533207909  2.53295926  2.532730420
#> hp           0.000000000  0.000000000  0.000000000  0.00000000  0.000000000
#> qsec         0.798254369  0.798235535  0.798218558  0.79820329  0.798189601
#> vs           1.480653565  1.480412520  1.480189821  1.47998428  1.479794765
#> wt          -2.300340299 -2.300476573 -2.300603309 -2.30072126 -2.300831113
#>                    0.0196       0.0183       0.0170       0.0159       0.0148
#> (Intercept) -6.443608e+00 -6.187166905 -6.184568555 -6.181928976 -6.179264189
#> am           3.372353e+00  3.399514682  3.399426750  3.399325555  3.399212632
#> carb        -1.455669e+00 -1.426055894 -1.425865914 -1.425682860 -1.425506761
#> cyl          8.054629e-01  0.830450447  0.830208389  0.829966352  0.829725366
#> disp         6.764713e-03  0.007055448  0.007059020  0.007062356  0.007065466
#> drat         1.748181e+00  1.729271880  1.729120530  1.728983065  1.728859019
#> gear         2.532520e+00  2.508033077  2.507776577  2.507516836  2.507255152
#> hp          -3.645505e-06 -0.003802324 -0.003798847 -0.003795440 -0.003792112
#> qsec         7.981232e-01  0.793483210  0.793498441  0.793509326  0.793516265
#> vs           1.479529e+00  1.438279560  1.438324100  1.438349286  1.438356767
#> wt          -2.300808e+00 -2.239136303 -2.239653732 -2.240142818 -2.240604376
#>                   0.0138       0.0129       0.0120       0.0112       0.0105
#> (Intercept) -6.176588784 -6.173915987 -6.171257722 -6.168624679 -6.166026382
#> am           3.399089446  3.398957385  3.398817766  3.398671827  3.398520729
#> carb        -1.425337618 -1.425175407 -1.425020083 -1.424871576 -1.424729801
#> cyl          0.829486360  0.829250171  0.829017543  0.828789136  0.828565528
#> disp         0.007068361  0.007071049  0.007073541  0.007075847  0.007077976
#> drat         1.728747881  1.728649101  1.728562103  1.728486289  1.728421050
#> gear         2.506992727  2.506730666  2.506469979  2.506211583  2.505956307
#> hp          -0.003788871 -0.003785722 -0.003782673 -0.003779727 -0.003776888
#> qsec         0.793519636  0.793519798  0.793517088  0.793511823  0.793504300
#> vs           1.438348133  1.438324908  1.438288551  1.438240449  1.438181921
#> wt          -2.241039258 -2.241448345 -2.241832540 -2.242192765 -2.242529951
#>                   0.0098      0.0091       0.0085       0.0079       0.0074
#> (Intercept) -6.163471252 -6.16096668 -6.158519067 -6.156133937 -6.153815948
#> am           3.398365558  3.39820733  3.398046967  3.397885346  3.397723256
#> carb        -1.424594652 -1.42446601 -1.424343739 -1.424227694 -1.424117719
#> cyl          0.828347222  0.82813465  0.827928174  0.827728096  0.827534661
#> disp         0.007079937  0.00708174  0.007083393  0.007084905  0.007086285
#> drat         1.728365767  1.72831982  1.728282605  1.728253503  1.728231922
#> gear         2.505704894  2.50545801  2.505216225  2.504980062  2.504749954
#> hp          -0.003774158 -0.00377154 -0.003769035 -0.003766642 -0.003764362
#> qsec         0.793494799  0.79348358  0.793470874  0.793456916  0.793441907
#> vs           1.438114212  1.43803850  1.437955886  1.437867406  1.437774024
#> wt          -2.242845038 -2.24313896 -2.243412667 -2.243667078 -2.243903119
#>                   0.0069       0.0064      0.0060       0.0056       0.0052
#> (Intercept) -6.151568979 -6.149396185 -6.14730005 -6.145282442 -6.143344669
#> am           3.397561421  3.397400499  3.39724108  3.397083710  3.396928854
#> carb        -1.424013646 -1.423915303 -1.42382251 -1.423735084 -1.423652834
#> cyl          0.827348058  0.827168428  0.82699587  0.826830427  0.826672126
#> disp         0.007087541  0.007088680  0.00708971  0.007090638  0.007091472
#> drat         1.728217278  1.728209007  1.72820656  1.728209414  1.728217060
#> gear         2.504526276  2.504309340  2.50409940  2.503896652  2.503701251
#> hp          -0.003762195 -0.003760138 -0.00375819 -0.003756350 -0.003754614
#> qsec         0.793426038  0.793409482  0.79339240  0.793374936  0.793357222
#> vs           1.437676637  1.437576076  1.43747311  1.437368435  1.437262706
#> wt          -2.244121696 -2.244323702 -2.24451001 -2.244681476 -2.244838929
#>                   0.0049       0.0045       0.0042       0.0039
#> (Intercept) -6.141487524 -6.139711337 -6.138016016 -6.136401093
#> am           3.396776934  3.396628320  3.396483330  3.396342237
#> carb        -1.423575571 -1.423503101 -1.423435231 -1.423371768
#> cyl          0.826520946  0.826376838  0.826239725  0.826109508
#> disp         0.007092217  0.007092881  0.007093469  0.007093988
#> drat         1.728229015  1.728244820  1.728264040  1.728286261
#> gear         2.503513302  2.503332865  2.503159965  2.502994591
#> hp          -0.003752981 -0.003751447 -0.003750009 -0.003748664
#> qsec         0.793339377  0.793321507  0.793303708  0.793286066
#> vs           1.437156505  1.437050367  1.436944772  1.436840151
#> wt          -2.244983179 -2.245115008 -2.245235174 -2.245344407
#> 
#> $iter
#>   [1]  0  6  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  1  4  3  3  3  3  3  3
#>  [26]  3  3  3 24 22 21 44  1  1  1  1  1  1  1  1 19 35  1  1 41  1  1  1  1  1
#>  [51]  1  1  1  1  1 91  1  1  1  1  1  1  1  1  1  1 91  1  1  1  1  1  1  1  1
#>  [76]  1  1 60  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> 
#> $lambda
#>   [1] 3.937126815 3.671777645 3.424312121 3.193524945 2.978292052 2.777565135
#>   [7] 2.590366541 2.415784506 2.252968716 2.101126165 1.959517294 1.827452386
#>  [13] 1.704288212 1.589424890 1.482302972 1.382400713 1.289231532 1.202341641
#>  [19] 1.121307839 1.045735443 0.975256374 0.909527359 0.848228259 0.791060513
#>  [25] 0.737745682 0.688024092 0.641653571 0.598408268 0.558077554 0.520464994
#>  [31] 0.485387396 0.452673910 0.422165203 0.393712682 0.367177765 0.342431212
#>  [37] 0.319352494 0.297829205 0.277756514 0.259036654 0.241578451 0.225296872
#>  [43] 0.210112616 0.195951729 0.182745239 0.170428821 0.158942489 0.148230297
#>  [49] 0.138240071 0.128923153 0.120234164 0.112130784 0.104573544 0.097525637
#>  [55] 0.090952736 0.084822826 0.079106051 0.073774568 0.068802409 0.064165357
#>  [61] 0.059840827 0.055807755 0.052046499 0.048538740 0.045267391 0.042216520
#>  [67] 0.039371268 0.036717776 0.034243121 0.031935249 0.029782921 0.027775651
#>  [73] 0.025903665 0.024157845 0.022529687 0.021011262 0.019595173 0.018274524
#>  [79] 0.017042882 0.015894249 0.014823030 0.013824007 0.012892315 0.012023416
#>  [85] 0.011213078 0.010457354 0.009752564 0.009095274 0.008482283 0.007910605
#>  [91] 0.007377457 0.006880241 0.006416536 0.005984083 0.005580776 0.005204650
#>  [97] 0.004853874 0.004526739 0.004221652 0.003937127
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 28
#> 
#> $loss
#>   [1] 476.77238 414.28232 362.00010 318.45672 282.38393 252.68743 228.42364
#>   [8] 208.77947 193.05486 180.64758 171.04000 163.78768 158.50945 154.87880
#>  [15] 152.61642 151.48375 151.25206 151.25206 145.90259 140.83529 136.61988
#>  [22] 133.13247 130.26620 127.92891 126.04120 124.53474 123.35075 122.43871
#>  [29] 118.42855 114.21074 111.22766  95.60157  95.60156  95.60156  95.60156
#>  [36]  95.60156  95.60156  95.60156  95.60156  95.60156  95.46129  85.11232
#>  [43]  85.11232  85.11231  80.50646  80.50646  80.50646  80.50645  80.50645
#>  [50]  80.50645  80.50645  80.50645  80.50645  80.50645  80.50645  76.77928
#>  [57]  76.77928  76.77927  76.77927  76.77926  76.77926  76.77926  76.77926
#>  [64]  76.77926  76.77925  76.77925  76.10752  76.10751  76.10750  76.10749
#>  [71]  76.10749  76.10748  76.10748  76.10748  76.10747  76.10747  76.10733
#>  [78]  76.03758  76.03757  76.03756  76.03755  76.03754  76.03753  76.03752
#>  [85]  76.03751  76.03750  76.03749  76.03749  76.03748  76.03748  76.03747
#>  [92]  76.03747  76.03746  76.03746  76.03746  76.03745  76.03745  76.03745
#>  [99]  76.03745  76.03745
#> 
#> $linear.predictors
#>      3.9371   3.6718   3.4243   3.1935   2.9783   2.7776   2.5904   2.4158
#> 1  19.65238 19.99827 20.32085 20.62169 20.90226 21.16391 21.40793 21.63551
#> 2  19.65238 19.87733 20.08712 20.28277 20.46523 20.63540 20.79409 20.94209
#> 3  19.65238 19.71607 19.77547 19.83086 19.88252 19.93070 19.97564 20.01754
#> 4  19.65238 19.60936 19.56923 19.53181 19.49691 19.46436 19.43401 19.40570
#> 5  19.65238 19.72793 19.79838 19.86409 19.92537 19.98252 20.03582 20.08552
#> 6  19.65238 19.74690 19.83505 19.91726 19.99392 20.06542 20.13211 20.19429
#> 7  19.65238 19.60936 19.56923 19.53181 19.49691 19.46436 19.43401 19.40570
#> 8  19.65238 19.60936 19.56923 19.53181 19.49691 19.46436 19.43401 19.40570
#> 9  19.65238 19.31055 18.99176 18.69446 18.41719 18.15861 17.91746 17.69256
#> 10 19.65238 19.47181 19.30341 19.14636 18.99990 18.86330 18.73592 18.61711
#> 11 19.65238 19.44810 19.25758 19.07991 18.91421 18.75967 18.61556 18.48115
#> 12 19.65238 18.75089 17.91016 17.12609 16.39486 15.71292 15.07694 14.48382
#> 13 19.65238 18.70583 17.82308 16.99982 16.23205 15.51602 14.84825 14.22549
#> 14 19.65238 20.37059 21.04040 21.66506 22.24762 22.79092 23.29761 23.77014
#> 15 19.65238 20.07179 20.46293 20.82771 21.16790 21.48517 21.78105 22.05700
#> 16 19.65238 19.57141 19.49590 19.42548 19.35980 19.29855 19.24143 19.18816
#> 17 19.65238 19.61173 19.57381 19.53846 19.50548 19.47473 19.44605 19.41930
#> 18 19.65238 19.41727 19.19800 18.99351 18.80281 18.62495 18.45909 18.30440
#> 19 19.65238 20.22593 20.76083 21.25968 21.72490 22.15877 22.56340 22.94076
#> 20 19.65238 19.92713 20.18336 20.42233 20.64518 20.85302 21.04685 21.22762
#> 21 19.65238 19.92239 20.17420 20.40903 20.62804 20.83229 21.02278 21.20043
#>      2.2530   2.1011   1.9595   1.8275   1.7043   1.5894   1.4823   1.3824
#> 1  21.84775 22.04568 22.23027 22.40242 22.56297 22.71270 22.85234 22.98256
#> 2  21.08012 21.20884 21.32889 21.44085 21.54526 21.64264 21.73345 21.81814
#> 3  20.05662 20.09307 20.12706 20.15875 20.18832 20.21589 20.24160 20.26558
#> 4  19.37930 19.35468 19.33172 19.31031 19.29034 19.27171 19.25435 19.23815
#> 5  20.13188 20.17511 20.21543 20.25303 20.28809 20.32080 20.35129 20.37974
#> 6  20.25229 20.30638 20.35682 20.40386 20.44773 20.48865 20.52681 20.56239
#> 7  19.37930 19.35468 19.33172 19.31031 19.29034 19.27171 19.25435 19.23815
#> 8  19.37930 19.35468 19.33172 19.31031 19.29034 19.27171 19.25435 19.23815
#> 9  17.48282 17.28721 17.10479 16.93466 16.77600 16.62803 16.49003 16.36134
#> 10 18.50632 18.40299 18.30663 18.21676 18.13294 18.05478 17.98188 17.91390
#> 11 18.35580 18.23891 18.12989 18.02821 17.93339 17.84496 17.76249 17.68558
#> 12 13.93067 13.41481 12.93371 12.48503 12.06660 11.67637 11.31243 10.97303
#> 13 13.64469 13.10304 12.59790 12.12680 11.68745 11.27772 10.89559 10.53922
#> 14 24.21083 24.62181 25.00510 25.36256 25.69592 26.00682 26.29676 26.56716
#> 15 22.31434 22.55434 22.77817 22.98691 23.18158 23.36313 23.53245 23.69035
#> 16 19.13848 19.09215 19.04894 19.00864 18.97106 18.93601 18.90332 18.87284
#> 17 19.39436 19.37109 19.34940 19.32916 19.31029 19.29270 19.27628 19.26098
#> 18 18.16013 18.02559 17.90012 17.78311 17.67398 17.57220 17.47729 17.38877
#> 19 23.29269 23.62089 23.92698 24.21244 24.47866 24.72694 24.95848 25.17442
#> 20 21.39620 21.55342 21.70005 21.83679 21.96432 22.08325 22.19417 22.29761
#> 21 21.36610 21.52061 21.66470 21.79908 21.92441 22.04129 22.15029 22.25195
#>      1.2892   1.2023   1.1213   1.0457   0.9753   0.9095   0.8482   0.7911
#> 1  23.07385 23.07385 22.94413 22.80785 22.68076 22.56223 22.45169 22.34860
#> 2  21.87751 21.87751 21.79190 21.70196 21.61809 21.53987 21.46692 21.39888
#> 3  20.28239 20.28239 20.37180 20.46574 20.55335 20.63505 20.71125 20.78231
#> 4  19.22679 19.22679 19.15275 19.07497 19.00243 18.93478 18.87169 18.81285
#> 5  20.39967 20.39967 20.52747 20.66172 20.78693 20.90370 21.01260 21.11416
#> 6  20.58734 20.58734 20.91570 21.26068 21.58241 21.88245 22.16227 22.42324
#> 7  19.22679 19.22679 19.24165 19.25725 19.27180 19.28538 19.29803 19.30984
#> 8  19.22679 19.22679 19.28331 19.34269 19.39807 19.44972 19.49788 19.54280
#> 9  16.27113 16.27113 16.23638 16.19988 16.16583 16.13408 16.10447 16.07685
#> 10 17.86625 17.86625 17.83843 17.80920 17.78194 17.75652 17.73281 17.71070
#> 11 17.63167 17.63167 17.63265 17.63369 17.63465 17.63554 17.63638 17.63716
#> 12 10.73513 10.73513 10.76482 10.79601 10.82510 10.85223 10.87754 10.90113
#> 13 10.28943 10.28943 10.28218 10.27456 10.26745 10.26083 10.25465 10.24888
#> 14 26.75669 26.75669 26.84980 26.94761 27.03883 27.12391 27.20325 27.27724
#> 15 23.80103 23.80103 23.91468 24.03407 24.14541 24.24925 24.34609 24.43641
#> 16 18.85147 18.85147 18.76865 18.68165 18.60051 18.52484 18.45426 18.38845
#> 17 19.25025 19.25025 19.19556 19.13809 19.08451 19.03453 18.98792 18.94445
#> 18 17.32672 17.32672 17.26306 17.19618 17.13381 17.07564 17.02139 16.97079
#> 19 25.32578 25.32578 25.20290 25.07381 24.95341 24.84113 24.73642 24.63877
#> 20 22.37012 22.37012 22.17680 21.97371 21.78431 21.60767 21.44293 21.28930
#> 21 22.32320 22.32320 22.34537 22.36867 22.39040 22.41066 22.42955 22.44717
#>      0.7377   0.6880   0.6417   0.5984   0.5581   0.5205   0.4854   0.4527
#> 1  22.25246 22.16280 22.07918 22.00119 22.00552 22.04221 22.07631 21.66513
#> 2  21.33544 21.27626 21.22108 21.16962 21.21785 21.30367 21.38356 21.07291
#> 3  20.84858 20.91038 20.96802 21.02178 20.86953 20.64161 20.42938 19.96181
#> 4  18.75798 18.70680 18.65908 18.61457 18.48541 18.32765 18.18068 18.63521
#> 5  21.20888 21.29721 21.37959 21.45642 21.51791 21.57103 21.62057 21.63577
#> 6  22.66661 22.89358 23.10526 23.30267 23.47704 23.63576 23.78377 22.75939
#> 7  19.32085 19.33112 19.34069 19.34962 19.50256 19.70665 19.89675 19.85038
#> 8  19.58469 19.62376 19.66020 19.69418 19.85529 20.06059 20.25183 19.85038
#> 9  16.05110 16.02708 16.00468 15.98379 15.92194 15.84623 15.77570 15.69582
#> 10 17.69008 17.67086 17.65292 17.63620 17.52870 17.38938 17.25961 16.48544
#> 11 17.63789 17.63857 17.63920 17.63979 17.54486 17.41576 17.29551 16.36932
#> 12 10.92314 10.94366 10.96280 10.98065 11.04646 11.12875 11.20541 11.21022
#> 13 10.24351 10.23849 10.23382 10.22946 10.40426 10.64329 10.86591 12.33401
#> 14 27.34625 27.41060 27.47062 27.52659 27.57611 27.62122 27.66328 28.27554
#> 15 24.52064 24.59919 24.67245 24.74077 24.70268 24.62399 24.55076 24.48209
#> 16 18.32707 18.26982 18.21644 18.16665 17.91278 17.58782 17.28510 16.70167
#> 17 18.90391 18.86611 18.83085 18.79797 18.67191 18.51378 18.36647 18.64682
#> 18 16.92361 16.87961 16.83857 16.80030 16.70510 16.59100 16.48469 17.38094
#> 19 24.54770 24.46276 24.38355 24.30968 24.43115 24.62525 24.80596 27.39054
#> 20 21.14602 21.01240 20.88779 20.77157 20.68792 20.62027 20.55716 17.82646
#> 21 22.46361 22.47894 22.49323 22.50656 22.63499 22.80408 22.96158 24.47015
#>      0.4222   0.3937   0.3672   0.3424   0.3194   0.2978   0.2778   0.2590
#> 1  21.66503 21.66497 21.66493 21.66491 21.66489 21.66488 21.66488 21.66488
#> 2  21.07287 21.07285 21.07284 21.07283 21.07283 21.07282 21.07282 21.07282
#> 3  19.96171 19.96165 19.96162 19.96160 19.96158 19.96157 19.96157 19.96157
#> 4  18.63511 18.63506 18.63502 18.63500 18.63498 18.63498 18.63497 18.63497
#> 5  21.63589 21.63595 21.63600 21.63602 21.63604 21.63605 21.63606 21.63606
#> 6  22.75961 22.75974 22.75983 22.75988 22.75991 22.75993 22.75994 22.75994
#> 7  19.85049 19.85056 19.85060 19.85063 19.85064 19.85065 19.85066 19.85066
#> 8  19.85049 19.85056 19.85060 19.85063 19.85064 19.85065 19.85066 19.85066
#> 9  15.69575 15.69571 15.69568 15.69567 15.69566 15.69565 15.69565 15.69565
#> 10 16.48529 16.48519 16.48514 16.48510 16.48508 16.48507 16.48506 16.48505
#> 11 16.36918 16.36909 16.36904 16.36901 16.36899 16.36898 16.36897 16.36897
#> 12 11.21028 11.21032 11.21035 11.21036 11.21037 11.21038 11.21038 11.21038
#> 13 12.33425 12.33440 12.33448 12.33454 12.33457 12.33459 12.33460 12.33461
#> 14 28.27568 28.27576 28.27581 28.27584 28.27586 28.27587 28.27588 28.27588
#> 15 24.48212 24.48214 24.48215 24.48215 24.48216 24.48216 24.48216 24.48216
#> 16 16.70140 16.70123 16.70113 16.70107 16.70103 16.70101 16.70099 16.70098
#> 17 18.64673 18.64667 18.64663 18.64661 18.64659 18.64659 18.64658 18.64658
#> 18 17.38090 17.38089 17.38087 17.38087 17.38086 17.38086 17.38086 17.38086
#> 19 27.39077 27.39091 27.39099 27.39104 27.39108 27.39109 27.39111 27.39111
#> 20 17.82608 17.82584 17.82570 17.82561 17.82556 17.82553 17.82551 17.82549
#> 21 24.47037 24.47051 24.47059 24.47064 24.47068 24.47070 24.47071 24.47071
#>      0.2416   0.2253   0.2101   0.1960   0.1827   0.1704   0.1589   0.1482
#> 1  21.67170 21.90011 21.90023 21.90032 21.21639 21.21640 21.21641 21.21642
#> 2  21.08198 21.71258 21.71278 21.71293 21.13511 21.13519 21.13526 21.13531
#> 3  19.96275 20.56227 20.56238 20.56246 20.82741 20.82761 20.82776 20.82788
#> 4  18.63342 17.86058 17.86042 17.86029 17.74803 17.74790 17.74781 17.74773
#> 5  21.62748 21.45387 21.45377 21.45370 22.25827 22.25827 22.25827 22.25827
#> 6  22.74898 23.80009 23.80022 23.80031 24.26543 24.26558 24.26570 24.26579
#> 7  19.83400 18.50766 18.50732 18.50707 18.33693 18.33657 18.33628 18.33605
#> 8  19.83400 18.85707 18.85680 18.85661 18.68379 18.68350 18.68326 18.68307
#> 9  15.69640 15.59748 15.59747 15.59746 15.46984 15.46984 15.46983 15.46983
#> 10 16.48269 16.39882 16.39881 16.39880 16.12549 16.12549 16.12549 16.12549
#> 11 16.36706 16.53104 16.53108 16.53111 16.27732 16.27736 16.27740 16.27742
#> 12 11.21901 12.19248 12.19272 12.19289 12.36402 12.36422 12.36438 12.36451
#> 13 12.34148 12.42740 12.42742 12.42744 12.23782 12.23775 12.23770 12.23766
#> 14 28.28524 29.33141 29.33162 29.33179 29.03648 29.03663 29.03676 29.03686
#> 15 24.47103 23.95990 23.95972 23.95958 23.06826 23.06803 23.06786 23.06771
#> 16 16.70359 16.63398 16.63399 16.63400 17.08129 17.08143 17.08154 17.08163
#> 17 18.64498 18.03371 18.03358 18.03348 17.91784 17.91774 17.91767 17.91761
#> 18 17.38364 16.88669 16.88659 16.88651 17.04080 17.04075 17.04071 17.04068
#> 19 27.39725 26.36545 26.36524 26.36509 27.30496 27.30481 27.30469 27.30460
#> 20 17.82779 18.30720 18.30742 18.30759 19.05260 19.05287 19.05309 19.05327
#> 21 24.48553 25.38020 25.38041 25.38057 25.25192 25.25205 25.25215 25.25223
#>      0.1382   0.1289   0.1202   0.1121   0.1046   0.0975   0.0910   0.0848
#> 1  21.21642 21.21643 21.21644 21.21645 21.21645 21.21646 21.21646 20.96488
#> 2  21.13536 21.13540 21.13543 21.13545 21.13548 21.13549 21.13551 20.97466
#> 3  20.82797 20.82805 20.82810 20.82815 20.82819 20.82821 20.82824 21.17378
#> 4  17.74766 17.74761 17.74757 17.74754 17.74751 17.74749 17.74747 17.97151
#> 5  22.25826 22.25825 22.25825 22.25824 22.25824 22.25824 22.25823 21.68630
#> 6  24.26587 24.26592 24.26597 24.26601 24.26604 24.26606 24.26608 24.23127
#> 7  18.33587 18.33572 18.33560 18.33551 18.33544 18.33538 18.33533 18.87161
#> 8  18.68292 18.68280 18.68270 18.68263 18.68257 18.68252 18.68249 19.31210
#> 9  15.46982 15.46982 15.46982 15.46982 15.46982 15.46982 15.46982 15.44363
#> 10 16.12549 16.12549 16.12549 16.12548 16.12548 16.12548 16.12548 16.12557
#> 11 16.27745 16.27746 16.27748 16.27749 16.27750 16.27751 16.27751 16.34053
#> 12 12.36462 12.36470 12.36477 12.36482 12.36486 12.36490 12.36492 12.10486
#> 13 12.23763 12.23761 12.23759 12.23758 12.23757 12.23756 12.23755 12.00468
#> 14 29.03694 29.03700 29.03706 29.03710 29.03713 29.03716 29.03718 29.78100
#> 15 23.06760 23.06751 23.06744 23.06739 23.06734 23.06731 23.06728 21.86538
#> 16 17.08169 17.08175 17.08179 17.08182 17.08184 17.08186 17.08188 17.13689
#> 17 17.91756 17.91752 17.91749 17.91747 17.91745 17.91743 17.91742 18.18494
#> 18 17.04065 17.04063 17.04061 17.04060 17.04058 17.04057 17.04057 17.24868
#> 19 27.30452 27.30445 27.30440 27.30436 27.30433 27.30430 27.30428 26.91339
#> 20 19.05341 19.05352 19.05360 19.05367 19.05373 19.05377 19.05380 18.88640
#> 21 25.25230 25.25235 25.25239 25.25243 25.25246 25.25248 25.25250 25.47795
#>      0.0791   0.0738   0.0688   0.0642   0.0598   0.0558   0.0520   0.0485
#> 1  20.96481 20.96475 20.96471 20.96468 20.96467 20.96466 20.96466 20.96466
#> 2  20.97458 20.97452 20.97446 20.97442 20.97439 20.97437 20.97435 20.97434
#> 3  21.17413 21.17442 21.17465 21.17484 21.17499 21.17511 21.17520 21.17527
#> 4  17.97139 17.97129 17.97119 17.97110 17.97102 17.97095 17.97088 17.97082
#> 5  21.68648 21.68665 21.68681 21.68695 21.68709 21.68722 21.68734 21.68745
#> 6  24.23102 24.23079 24.23060 24.23042 24.23027 24.23013 24.23002 24.22992
#> 7  18.87141 18.87123 18.87107 18.87094 18.87082 18.87072 18.87063 18.87055
#> 8  19.31184 19.31162 19.31141 19.31123 19.31108 19.31094 19.31081 19.31070
#> 9  15.44359 15.44356 15.44354 15.44352 15.44350 15.44348 15.44347 15.44346
#> 10 16.12546 16.12537 16.12529 16.12522 16.12516 16.12510 16.12506 16.12501
#> 11 16.34040 16.34028 16.34018 16.34008 16.34000 16.33992 16.33986 16.33979
#> 12 12.10502 12.10517 12.10530 12.10542 12.10552 12.10561 12.10569 12.10576
#> 13 12.00467 12.00467 12.00468 12.00470 12.00473 12.00476 12.00479 12.00483
#> 14 29.78100 29.78098 29.78094 29.78090 29.78084 29.78077 29.78070 29.78062
#> 15 21.86535 21.86536 21.86540 21.86547 21.86556 21.86567 21.86579 21.86592
#> 16 17.13711 17.13729 17.13744 17.13756 17.13765 17.13773 17.13779 17.13783
#> 17 18.18480 18.18467 18.18455 18.18444 18.18434 18.18425 18.18417 18.18409
#> 18 17.24868 17.24868 17.24867 17.24865 17.24863 17.24861 17.24859 17.24857
#> 19 26.91332 26.91326 26.91323 26.91320 26.91319 26.91319 26.91320 26.91321
#> 20 18.88669 18.88694 18.88715 18.88733 18.88748 18.88761 18.88771 18.88780
#> 21 25.47825 25.47850 25.47872 25.47891 25.47906 25.47919 25.47930 25.47939
#>      0.0453   0.0422   0.0394   0.0367   0.0342   0.0319   0.0298   0.0278
#> 1  20.96467 20.96469 21.07657 21.07672 21.07686 21.07699 21.07710 21.07721
#> 2  20.97433 20.97433 20.93730 20.93738 20.93746 20.93753 20.93759 20.93765
#> 3  21.17531 21.17534 21.25969 21.25951 21.25934 21.25918 21.25904 21.25890
#> 4  17.97077 17.97072 18.29690 18.29692 18.29693 18.29695 18.29696 18.29698
#> 5  21.68756 21.68765 21.54188 21.54194 21.54199 21.54203 21.54207 21.54211
#> 6  24.22983 24.22975 24.31094 24.31097 24.31099 24.31102 24.31104 24.31106
#> 7  18.87049 18.87043 18.85319 18.85318 18.85317 18.85317 18.85316 18.85316
#> 8  19.31061 19.31053 19.33222 19.33219 19.33216 19.33214 19.33213 19.33211
#> 9  15.44345 15.44344 14.98614 14.98607 14.98601 14.98595 14.98590 14.98585
#> 10 16.12498 16.12495 15.92764 15.92763 15.92763 15.92763 15.92762 15.92762
#> 11 16.33974 16.33969 16.13202 16.13199 16.13197 16.13194 16.13192 16.13191
#> 12 12.10582 12.10587 12.36000 12.36005 12.36009 12.36013 12.36017 12.36020
#> 13 12.00487 12.00491 12.00213 12.00221 12.00229 12.00236 12.00243 12.00249
#> 14 29.78055 29.78047 29.82478 29.82457 29.82439 29.82422 29.82406 29.82391
#> 15 21.86605 21.86618 21.97599 21.97643 21.97683 21.97720 21.97754 21.97786
#> 16 17.13785 17.13787 17.02839 17.02825 17.02812 17.02799 17.02788 17.02777
#> 17 18.18402 18.18396 18.15360 18.15354 18.15349 18.15344 18.15340 18.15336
#> 18 17.24854 17.24852 17.53757 17.53754 17.53750 17.53747 17.53745 17.53742
#> 19 26.91323 26.91325 26.86429 26.86433 26.86436 26.86439 26.86442 26.86444
#> 20 18.88787 18.88793 18.99676 18.99672 18.99669 18.99666 18.99663 18.99660
#> 21 25.47946 25.47952 25.30201 25.30186 25.30173 25.30160 25.30148 25.30137
#>      0.0259   0.0242   0.0225   0.0210   0.0196   0.0183   0.0170   0.0159
#> 1  21.07731 21.07740 21.07749 21.07757 21.07777 21.17193 21.17198 21.17203
#> 2  20.93770 20.93775 20.93780 20.93784 20.93802 21.04530 21.04523 21.04516
#> 3  21.25877 21.25865 21.25854 21.25844 21.25830 21.28656 21.28664 21.28671
#> 4  18.29699 18.29701 18.29702 18.29704 18.29702 18.25264 18.25281 18.25297
#> 5  21.54214 21.54217 21.54220 21.54222 21.54234 21.66006 21.65992 21.65980
#> 6  24.31109 24.31111 24.31113 24.31115 24.31098 24.28135 24.28134 24.28133
#> 7  18.85316 18.85316 18.85316 18.85316 18.85316 18.87339 18.87325 18.87311
#> 8  19.33210 19.33209 19.33208 19.33208 19.33204 19.34948 19.34935 19.34921
#> 9  14.98581 14.98577 14.98573 14.98569 14.98567 14.96603 14.96580 14.96558
#> 10 15.92762 15.92762 15.92761 15.92761 15.92757 15.88603 15.88598 15.88593
#> 11 16.13189 16.13187 16.13186 16.13185 16.13178 16.09147 16.09140 16.09133
#> 12 12.36023 12.36025 12.36028 12.36030 12.36036 12.40514 12.40530 12.40546
#> 13 12.00255 12.00260 12.00265 12.00270 12.00272 11.94602 11.94605 11.94609
#> 14 29.82378 29.82365 29.82354 29.82343 29.82326 29.81202 29.81197 29.81191
#> 15 21.97815 21.97841 21.97866 21.97889 21.97897 21.90593 21.90626 21.90659
#> 16 17.02767 17.02758 17.02750 17.02742 17.02742 17.07880 17.07875 17.07869
#> 17 18.15333 18.15330 18.15327 18.15324 18.15327 18.18596 18.18585 18.18574
#> 18 17.53740 17.53738 17.53736 17.53735 17.53735 17.53076 17.53088 17.53098
#> 19 26.86447 26.86449 26.86452 26.86454 26.86469 26.84505 26.84492 26.84479
#> 20 18.99657 18.99655 18.99652 18.99649 18.99644 18.89306 18.89336 18.89364
#> 21 25.30127 25.30117 25.30108 25.30099 25.30087 25.23300 25.23298 25.23295
#>      0.0148   0.0138   0.0129   0.0120   0.0112   0.0105   0.0098   0.0091
#> 1  21.17208 21.17213 21.17217 21.17222 21.17226 21.17231 21.17235 21.17239
#> 2  21.04510 21.04503 21.04498 21.04492 21.04487 21.04482 21.04478 21.04474
#> 3  21.28676 21.28680 21.28684 21.28686 21.28687 21.28687 21.28687 21.28686
#> 4  18.25311 18.25325 18.25337 18.25349 18.25359 18.25369 18.25378 18.25386
#> 5  21.65968 21.65958 21.65948 21.65939 21.65932 21.65924 21.65918 21.65912
#> 6  24.28131 24.28130 24.28129 24.28127 24.28126 24.28125 24.28124 24.28122
#> 7  18.87298 18.87285 18.87273 18.87261 18.87251 18.87240 18.87230 18.87221
#> 8  19.34909 19.34896 19.34884 19.34872 19.34861 19.34851 19.34840 19.34830
#> 9  14.96538 14.96519 14.96502 14.96486 14.96471 14.96457 14.96445 14.96433
#> 10 15.88589 15.88585 15.88582 15.88578 15.88576 15.88573 15.88571 15.88570
#> 11 16.09126 16.09121 16.09115 16.09110 16.09105 16.09101 16.09097 16.09093
#> 12 12.40560 12.40574 12.40586 12.40598 12.40609 12.40619 12.40628 12.40637
#> 13 11.94613 11.94618 11.94623 11.94628 11.94634 11.94639 11.94645 11.94651
#> 14 29.81184 29.81176 29.81168 29.81159 29.81150 29.81141 29.81131 29.81122
#> 15 21.90693 21.90727 21.90761 21.90794 21.90828 21.90861 21.90893 21.90925
#> 16 17.07864 17.07858 17.07851 17.07845 17.07839 17.07832 17.07826 17.07820
#> 17 18.18564 18.18553 18.18543 18.18533 18.18523 18.18514 18.18505 18.18496
#> 18 17.53107 17.53115 17.53123 17.53129 17.53135 17.53140 17.53144 17.53148
#> 19 26.84468 26.84458 26.84449 26.84441 26.84434 26.84428 26.84422 26.84418
#> 20 18.89391 18.89416 18.89440 18.89462 18.89484 18.89503 18.89522 18.89539
#> 21 25.23293 25.23290 25.23288 25.23286 25.23284 25.23282 25.23279 25.23277
#>      0.0085   0.0079   0.0074   0.0069   0.0064   0.0060   0.0056   0.0052
#> 1  21.17243 21.17247 21.17250 21.17254 21.17257 21.17260 21.17264 21.17267
#> 2  21.04470 21.04467 21.04463 21.04461 21.04458 21.04455 21.04453 21.04451
#> 3  21.28684 21.28682 21.28680 21.28677 21.28673 21.28669 21.28665 21.28661
#> 4  18.25393 18.25400 18.25406 18.25411 18.25416 18.25421 18.25425 18.25428
#> 5  21.65907 21.65903 21.65899 21.65896 21.65893 21.65891 21.65888 21.65887
#> 6  24.28121 24.28120 24.28119 24.28118 24.28117 24.28116 24.28115 24.28115
#> 7  18.87213 18.87205 18.87197 18.87190 18.87183 18.87177 18.87171 18.87166
#> 8  19.34821 19.34812 19.34803 19.34795 19.34788 19.34781 19.34774 19.34767
#> 9  14.96423 14.96413 14.96405 14.96397 14.96390 14.96383 14.96377 14.96372
#> 10 15.88568 15.88567 15.88566 15.88565 15.88565 15.88564 15.88564 15.88564
#> 11 16.09090 16.09087 16.09084 16.09082 16.09080 16.09077 16.09076 16.09074
#> 12 12.40645 12.40652 12.40659 12.40665 12.40670 12.40675 12.40680 12.40684
#> 13 11.94657 11.94663 11.94669 11.94675 11.94680 11.94686 11.94692 11.94697
#> 14 29.81112 29.81102 29.81092 29.81083 29.81073 29.81063 29.81054 29.81045
#> 15 21.90956 21.90986 21.91016 21.91044 21.91072 21.91099 21.91125 21.91150
#> 16 17.07813 17.07807 17.07801 17.07794 17.07788 17.07782 17.07777 17.07771
#> 17 18.18488 18.18480 18.18472 18.18465 18.18458 18.18451 18.18445 18.18439
#> 18 17.53151 17.53154 17.53156 17.53157 17.53159 17.53160 17.53160 17.53161
#> 19 26.84414 26.84410 26.84408 26.84406 26.84404 26.84403 26.84402 26.84402
#> 20 18.89556 18.89571 18.89584 18.89597 18.89609 18.89620 18.89630 18.89639
#> 21 25.23275 25.23273 25.23271 25.23269 25.23267 25.23265 25.23263 25.23261
#>      0.0049   0.0045   0.0042   0.0039
#> 1  21.17270 21.17272 21.17275 21.17278
#> 2  21.04450 21.04448 21.04447 21.04445
#> 3  21.28657 21.28652 21.28647 21.28642
#> 4  18.25431 18.25434 18.25436 18.25438
#> 5  21.65885 21.65884 21.65884 21.65883
#> 6  24.28114 24.28113 24.28113 24.28112
#> 7  18.87161 18.87157 18.87152 18.87149
#> 8  19.34761 19.34756 19.34751 19.34746
#> 9  14.96368 14.96364 14.96360 14.96357
#> 10 15.88564 15.88564 15.88564 15.88564
#> 11 16.09073 16.09071 16.09070 16.09069
#> 12 12.40687 12.40691 12.40694 12.40696
#> 13 11.94702 11.94708 11.94713 11.94718
#> 14 29.81036 29.81027 29.81018 29.81010
#> 15 21.91174 21.91197 21.91219 21.91240
#> 16 17.07765 17.07760 17.07755 17.07750
#> 17 18.18433 18.18428 18.18423 18.18418
#> 18 17.53161 17.53160 17.53160 17.53159
#> 19 26.84402 26.84402 26.84403 26.84404
#> 20 18.89647 18.89655 18.89662 18.89668
#> 21 25.23260 25.23258 25.23256 25.23254
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 21.4 18.7 24.4 22.8 19.2 17.8 16.4 17.3 15.2 10.4 14.7 33.9 21.5
#> [16] 15.5 15.2 19.2 26.0 19.7 21.4
#> 
#> $X
#>               am       carb        cyl       disp       drat       gear
#>  [1,]  1.5811388  0.9757626 -0.1714986 -0.6471940  0.7961051  0.6495191
#>  [2,]  1.5811388  0.9757626 -0.1714986 -0.6471940  0.7961051  0.6495191
#>  [3,] -0.6324555 -1.3885853 -0.1714986  0.2179172 -0.9218059 -0.8660254
#>  [4,] -0.6324555 -0.6004693  1.0289915  1.1183392 -0.7751550 -0.8660254
#>  [5,] -0.6324555 -0.6004693 -1.3719887 -0.7646020  0.3561523  0.6495191
#>  [6,] -0.6324555 -0.6004693 -1.3719887 -0.8166852  0.8380054  0.6495191
#>  [7,] -0.6324555  0.9757626 -0.1714986 -0.5801038  0.8380054  0.6495191
#>  [8,] -0.6324555  0.9757626 -0.1714986 -0.5801038  0.8380054  0.6495191
#>  [9,] -0.6324555  0.1876467  1.0289915  0.3750497 -0.9427561 -0.8660254
#> [10,] -0.6324555  0.1876467  1.0289915  0.3750497 -0.9427561 -0.8660254
#> [11,] -0.6324555  0.1876467  1.0289915  0.3750497 -0.9427561 -0.8660254
#> [12,] -0.6324555  0.9757626  1.0289915  2.1070377 -1.2360580 -0.8660254
#> [13,] -0.6324555  0.9757626  1.0289915  1.8245524 -0.6075539 -0.8660254
#> [14,]  1.5811388 -1.3885853 -1.3719887 -1.4319735  1.4665095  0.6495191
#> [15,] -0.6324555 -1.3885853 -1.3719887 -0.9994179  0.3771024 -0.8660254
#> [16,] -0.6324555 -0.6004693  1.0289915  0.7475772 -1.5922103 -0.8660254
#> [17,] -0.6324555 -0.6004693  1.0289915  0.6239899 -0.7751550 -0.8660254
#> [18,] -0.6324555 -0.6004693  1.0289915  1.4714458 -0.9218059 -0.8660254
#> [19,]  1.5811388 -0.6004693 -1.3719887 -0.9976524  1.9064623  2.1650635
#> [20,]  1.5811388  2.5519945 -0.1714986 -0.7796090  0.2095014  2.1650635
#> [21,]  1.5811388 -0.6004693 -1.3719887 -0.9914730  1.2360580  0.6495191
#>               hp        qsec         vs         wt
#>  [1,] -0.6147149 -1.02614131 -0.7844645 -0.8690257
#>  [2,] -0.6147149 -0.67847718 -0.7844645 -0.5651646
#>  [3,] -0.6147149  0.82392850  1.2747549 -0.1600165
#>  [4,]  0.8196199 -0.67847718 -0.7844645  0.1080962
#>  [5,] -1.6739160  1.17159263  1.2747549 -0.1898068
#>  [6,] -0.9457153  2.97199614  1.2747549 -0.2374713
#>  [7,] -0.3278480  0.11618367  1.2747549  0.1080962
#>  [8,] -0.3278480  0.48868095  1.2747549  0.1080962
#>  [9,]  0.9299533 -0.44256224 -0.7844645  0.8588118
#> [10,]  0.9299533 -0.31839648 -0.7844645  0.4536637
#> [11,]  0.9299533 -0.07006496 -0.7844645  0.5132443
#> [12,]  1.4816206 -0.08248154 -0.7844645  2.2649141
#> [13,]  2.0332878 -0.43014567 -0.7844645  2.3781172
#> [14,] -1.6077159  1.10950975  1.2747549 -1.8044411
#> [15,] -0.9015819  1.17780092  1.2747549 -1.0537255
#> [16,]  0.2679527 -0.77160150 -0.7844645  0.2034252
#> [17,]  0.2679527 -0.50464512 -0.7844645  0.1021382
#> [18,]  0.8196199 -0.65985232 -0.7844645  0.5906991
#> [19,] -1.0339820 -0.87714240 -0.7844645 -1.4409995
#> [20,]  0.8196199 -1.62213696 -0.7844645 -0.6902839
#> [21,] -0.6367816  0.30243231  1.2747549 -0.6783677
#> attr(,"center")
#>  [1]   0.2857143   2.7619048   6.2857143 233.3142857   3.5200000   3.5714286
#>  [7] 137.8571429  18.1128571   0.3809524   3.3492857
#> attr(,"scale")
#>  [1]   0.4517540   1.2688488   1.6659863 113.2802269   0.4773239   0.6598289
#>  [7]  45.3171740   1.6107500   0.4856209   0.8391993
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 27
#> 
#> $lambda.min
#> [1] 0.6416536
#> 
#> $null.dev
#> [1] 22.70345
#> 
#> $Bias
#> [1] 4.462004
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 9.034952