Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 36.09695 34.48241 32.97161 30.86301 28.75913 27.16136 25.98815 25.16960
#>   [9] 24.64592 24.36601 24.20420 24.14590 24.23056 24.43027 24.72157 24.65044
#>  [17] 24.10822 23.53635 22.95055 22.40082 22.00629 21.62993 21.32851 21.04899
#>  [25] 20.45666 19.92296 19.41056 18.46093 17.99876 17.67832 17.54501 17.56619
#>  [33] 16.93640 17.11120 17.28466 20.82971 20.31789 20.25976 20.54576 20.25936
#>  [41] 19.83842 18.84597 17.13494 16.76271 11.87664 12.04256 11.90814 11.75035
#>  [49] 11.75038 11.62593 11.62612 12.76901 12.76979 13.54045 13.54144 15.76467
#>  [57] 20.34008 20.84212 20.81238 20.84748 20.84886 21.01977 21.02120 21.07852
#>  [65] 20.60203 20.38703 20.38867 20.39029 20.39187 20.40028 20.40177 20.40322
#>  [73] 20.40325 20.40455 20.40584 20.95815 20.96044 20.95136 20.74039 20.74234
#>  [81] 20.74425 20.74611 20.74793 20.74969 20.75141 20.75309 20.78519 20.70746
#>  [89] 20.72296 20.80209 20.80397 20.80580 20.80757 20.80930 20.81028 20.81191
#>  [97] 20.81349 20.81503 20.81652 20.81797
#> 
#> $cvse
#>   [1] 11.730894 11.783629 11.656063 11.059455 10.244348  9.494684  8.820086
#>   [8]  8.234372  7.753373  7.392237  7.161890  7.067212  7.104276  7.260367
#>  [15]  7.516992  7.574162  7.409722  7.235071  7.126362  7.096203  7.080044
#>  [22]  7.069812  7.057949  7.037996  7.004929  6.891039  6.772081  6.329930
#>  [29]  6.297923  6.254933  6.234302  6.234478  6.168958  6.164208  6.161226
#>  [36]  7.026880  6.907080  6.914451  6.890120  6.911000  6.913495  6.725985
#>  [43]  6.751475  6.742056  4.351609  4.373391  4.368837  4.367936  4.368038
#>  [50]  4.372165  4.372229  5.008611  5.009057  4.988617  4.989013  5.225835
#>  [57]  7.816147  7.814379  7.819369  7.791796  7.792874  7.774824  7.775853
#>  [64]  7.783065  7.819183  7.825903  7.826830  7.827733  7.828611  7.814619
#>  [71]  7.815438  7.816236  7.816678  7.817438  7.818178  7.996337  7.997479
#>  [78]  7.999659  7.996243  7.997291  7.998315  7.999314  8.000289  8.001240
#>  [85]  8.002167  8.003071  8.001071  8.007627  8.007242  8.010579  8.011383
#>  [92]  8.012166  8.012928  8.013669  8.012678  8.013381  8.014065  8.014730
#>  [99]  8.015378  8.016008
#> 
#> $fold
#>  [1]  3  3  9  1  2  7  4  8  8  6 10 10  1  7  1  9  5  4  2  5  6
#> 
#> $lambda
#>   [1] 4.724057219 4.405671573 4.108744054 3.831828456 3.573576043 3.332728978
#>   [7] 3.108114199 2.898637705 2.703279225 2.521087253 2.351174410 2.192713124
#>  [13] 2.044931598 1.907110052 1.778577218 1.658707066 1.546915761 1.442658816
#>  [19] 1.345428440 1.254751066 1.170185044 1.091318489 1.017767275 0.949173167
#>  [25] 0.885202072 0.825542415 0.769903619 0.718014692 0.669622905 0.624492562
#>  [31] 0.582403854 0.543151783 0.506545171 0.472405722 0.440567157 0.410874405
#>  [37] 0.383182846 0.357357604 0.333272898 0.310811420 0.289863770 0.270327922
#>  [43] 0.252108725 0.235117441 0.219271312 0.204493160 0.190711005 0.177857722
#>  [49] 0.165870707 0.154691576 0.144265882 0.134542844 0.125475107 0.117018504
#>  [55] 0.109131849 0.101776727 0.094917317 0.088520207 0.082554242 0.076990362
#>  [61] 0.071801469 0.066962291 0.062449256 0.058240385 0.054315178 0.050654517
#>  [67] 0.047240572 0.044056716 0.041087441 0.038318285 0.035735760 0.033327290
#>  [73] 0.031081142 0.028986377 0.027032792 0.025210873 0.023511744 0.021927131
#>  [79] 0.020449316 0.019071101 0.017785772 0.016587071 0.015469158 0.014426588
#>  [85] 0.013454284 0.012547511 0.011701850 0.010913185 0.010177673 0.009491732
#>  [91] 0.008852021 0.008255424 0.007699036 0.007180147 0.006696229 0.006244926
#>  [97] 0.005824039 0.005431518 0.005065452 0.004724057
#> 
#> $fit
#> $beta
#>               4.7241     4.4057     4.1087     3.8318     3.5736    3.3327
#> (Intercept) 20.76667 22.3359864 23.7995392 25.1644534 26.4373770 27.624510
#> am           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> cyl          0.00000 -0.2657719 -0.5136316 -0.7447865 -0.9603622 -1.161409
#> disp         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> wt           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#>                3.1081    2.8986    2.7033    2.5211   2.3512    2.1927
#> (Intercept) 28.731634 29.764141 30.727061 31.625084 32.46258 33.243637
#> am           0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> cyl         -1.348906 -1.523766 -1.686841 -1.838926 -1.98076 -2.113035
#> disp         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#>                2.0449    1.9071    1.7786    1.6587    1.5469    1.4427
#> (Intercept) 33.972050 34.651371 35.284909 35.875747 36.289888 36.289888
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.236396 -2.351442 -2.458735 -2.558796 -2.628933 -2.628933
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.3454    1.2548      1.1702     1.0913     1.0178    0.9492
#> (Intercept) 36.289888 36.289888 36.26333861 35.7966552 35.3614123 34.955503
#> am           0.000000  0.000000  0.02239682  0.4160918  0.7832636  1.125689
#> carb         0.000000  0.000000  0.00000000  0.0000000  0.0000000  0.000000
#> cyl         -2.628933 -2.628933 -2.62588133 -2.5722459 -2.5222239 -2.475573
#> disp         0.000000  0.000000  0.00000000  0.0000000  0.0000000  0.000000
#> drat         0.000000  0.000000  0.00000000  0.0000000  0.0000000  0.000000
#> gear         0.000000  0.000000  0.00000000  0.0000000  0.0000000  0.000000
#> hp           0.000000  0.000000  0.00000000  0.0000000  0.0000000  0.000000
#> qsec         0.000000  0.000000  0.00000000  0.0000000  0.0000000  0.000000
#> vs           0.000000  0.000000  0.00000000  0.0000000  0.0000000  0.000000
#> wt           0.000000  0.000000  0.00000000  0.0000000  0.0000000  0.000000
#>                0.8852    0.8255    0.7699    0.7180    0.6696    0.6245
#> (Intercept) 39.679367 39.744822 39.744742 39.744697 39.744671 39.744656
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.536850 -1.515449 -1.515342 -1.515280 -1.515245 -1.515225
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -3.175246 -3.237157 -3.237335 -3.237438 -3.237497 -3.237531
#>                0.5824    0.5432    0.5065    0.4724    0.4406      0.4109
#> (Intercept) 39.744648 39.744643 39.744640 39.744638 39.744637 39.74405065
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000 -0.01044078
#> cyl         -1.515213 -1.515206 -1.515202 -1.515200 -1.515199 -1.51222381
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> wt          -3.237550 -3.237561 -3.237568 -3.237571 -3.237573 -3.23454930
#>                  0.3832      0.3574      0.3333      0.3108      0.2899
#> (Intercept) 39.74262167 20.45653364 21.58020586 22.29898092 22.29791579
#> am           0.00000000  3.75668102  4.32727464  4.69598709  4.69543094
#> carb        -0.06285647  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl         -1.49934236  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         0.00000000  1.94295262  2.38610441  2.67311549  2.67325984
#> gear         0.00000000  0.00000000 -0.83551182 -1.37426573 -1.37405839
#> hp           0.00000000 -0.06154692 -0.06061034 -0.06000015 -0.06000034
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -3.21594113  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.2703      0.2521      0.2351      0.2193      0.2045
#> (Intercept) 22.29742570  2.25809198  2.25413320  2.25131173  2.24931163
#> am           4.69508887  5.58981684  5.59011851  5.59033580  5.59049139
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         2.67327316  3.39811701  3.39819258  3.39824770  3.39828758
#> gear        -1.37389338 -1.38911235 -1.38909396 -1.38908205 -1.38907437
#> hp          -0.06000065 -0.05381391 -0.05380812 -0.05380391 -0.05380086
#> qsec         0.00000000  0.98652058  0.98666950  0.98677460  0.98684836
#> vs           0.00000000 -3.22314327 -3.22312541 -3.22310187 -3.22307737
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.1907      0.1779      0.1659      0.1547     0.1443
#> (Intercept)  2.24790171  2.24691365  2.24622550  2.24574942  2.2454224
#> am           5.59060215  5.59068056  5.59073576  5.59077440  5.5908013
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> drat         3.39831620  3.39833657  3.39835094  3.39836098  3.3983679
#> gear        -1.38906943 -1.38906629 -1.38906430 -1.38906308 -1.3890623
#> hp          -0.05379866 -0.05379708 -0.05379596 -0.05379516 -0.0537946
#> qsec         0.98689981  0.98693549  0.98696005  0.98697683  0.9869882
#> vs          -3.22305443 -3.22303419 -3.22301700 -3.22300280 -3.2229913
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#>                  0.1345      0.1255      0.1170      0.1091      0.1018
#> (Intercept)  2.24519957  2.24504905  2.24494840  2.24488187 -6.61396750
#> am           5.59081991  5.59083271  5.59084145  5.59084738  6.07990107
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.66828521
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         3.39837270  3.39837594  3.39837811  3.39837954  3.79092183
#> gear        -1.38906190 -1.38906167 -1.38906156 -1.38906152 -1.14228904
#> hp          -0.05379421 -0.05379394 -0.05379375 -0.05379362 -0.06195345
#> qsec         0.98699584  0.98700091  0.98700424  0.98700638  1.16908481
#> vs          -3.22298217 -3.22297498 -3.22296940 -3.22296509 -2.93588521
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.0949      0.0885      0.0826      0.0770      0.0718
#> (Intercept) -6.61498766 -6.61516778 -6.61472389 -6.61383113 -6.61263006
#> am           6.08074396  6.08138647  6.08186493  6.08221030  6.08244878
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.66839035  0.66842940  0.66841872  0.66837167  0.66829899
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         3.79057530  3.79027989  3.79002966  3.78981910  3.78964312
#> gear        -1.14238162 -1.14247867 -1.14257655 -1.14267255 -1.14276474
#> hp          -0.06194856 -0.06194407 -0.06193999 -0.06193631 -0.06193303
#> qsec         1.16912182  1.16913616  1.16913343  1.16911817  1.16909407
#> vs          -2.93517989 -2.93462640 -2.93419889 -2.93387509 -2.93363597
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.0670      0.0624      0.0582      0.0543      0.0507
#> (Intercept) -6.61123216 -6.60972466 -6.60817463 -6.16063842 -1.97056823
#> am           6.08260240  6.08268956  6.08272552  5.80258785  6.42353095
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.99445222
#> cyl          0.66820928  0.66810929  0.66800427  0.67255795  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.02523107
#> drat         3.78949711  3.78937690  3.78927877  3.65323033  2.79496186
#> gear        -1.14285176 -1.14293278 -1.14300731 -1.09456693 -1.61805074
#> hp          -0.06193012 -0.06192756 -0.06192533 -0.05790899 -0.07159323
#> qsec         1.16906404  1.16903042  1.16899500  1.20555942  1.44464776
#> vs          -2.93346535 -2.93334955 -2.93327712 -2.88305896 -2.43363974
#> wt           0.00000000  0.00000000  0.00000000 -0.41234797 -2.29426089
#>                  0.0472      0.0441      0.0411      0.0383      0.0357
#> (Intercept) -1.97415110 -1.97718073 -1.97974069 -1.98190536 -1.98373959
#> am           6.42403869  6.42447537  6.42485283  6.42518099  6.42546810
#> carb         0.99467586  0.99487537  0.99505451  0.99521631  0.99536321
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.02523485  0.02523841  0.02524176  0.02524490  0.02524783
#> drat         2.79491036  2.79483915  2.79475366  2.79465844  2.79455720
#> gear        -1.61824881 -1.61841191 -1.61854775 -1.61866230 -1.61876016
#> hp          -0.07159207 -0.07159107 -0.07159021 -0.07158946 -0.07158882
#> qsec         1.44488000  1.44508205  1.44525840  1.44541294  1.44554896
#> vs          -2.43369252 -2.43370607 -2.43368979 -2.43365139 -2.43359712
#> wt          -2.29470816 -2.29513179 -2.29553257 -2.29591115 -2.29626809
#>                  0.0333      0.0311      0.0290      0.0270      0.0252
#> (Intercept) -1.98529889 -1.98663019 -1.98777273 -1.98875907 -1.98961598
#> am           6.42572094  6.42594507  6.42614501  6.42632447  6.42648643
#> carb         0.99549717  0.99561982  0.99573246  0.99583619  0.99593192
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.02525058  0.02525313  0.02525552  0.02525774  0.02525981
#> drat         2.79445293  2.79434797  2.79424410  2.79414266  2.79404461
#> gear        -1.61884485 -1.61891905 -1.61898482 -1.61904371 -1.61909692
#> hp          -0.07158826 -0.07158778 -0.07158737 -0.07158700 -0.07158669
#> qsec         1.44566925  1.44577615  1.44587161  1.44595727  1.44603447
#> vs          -2.43353194 -2.43345978 -2.43338368 -2.43330594 -2.43322830
#> wt          -2.29660403 -2.29691963 -2.29721562 -2.29749278 -2.29775196
#>                  0.0235      0.0219      0.0204      0.0191      0.0178
#> (Intercept) -1.99036542 -1.99102529 -1.99161015 -1.99213182 -1.99259990
#> am           6.42663334  6.42676719  6.42688961  6.42700195  6.42710533
#> carb         0.99602044  0.99610239  0.99617836  0.99624883  0.99631426
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.02526174  0.02526353  0.02526519  0.02526674  0.02526818
#> drat         2.79395061  2.79386107  2.79377624  2.79369621  2.79362096
#> gear        -1.61914534 -1.61918969 -1.61923050 -1.61926819 -1.61930310
#> hp          -0.07158641 -0.07158617 -0.07158596 -0.07158578 -0.07158562
#> qsec         1.44610433  1.44616781  1.44622567  1.44627858  1.44632710
#> vs          -2.43315203 -2.43307804 -2.43300693 -2.43293910 -2.43287478
#> wt          -2.29799402 -2.29821986 -2.29843038 -2.29862647 -2.29880899
#>                  0.0166      0.0155      0.0144      0.0135      0.0125
#> (Intercept) -1.99302220 -1.99340507 -1.99375369 -1.99407235 -1.99436456
#> am           6.42720067  6.42728877  6.42737032  6.42744588  6.42751598
#> carb         0.99637504  0.99643152  0.99648402  0.99653284  0.99657824
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.02526952  0.02527077  0.02527193  0.02527300  0.02527401
#> drat         2.79355041  2.79348441  2.79342277  2.79336529  2.79331174
#> gear        -1.61933551 -1.61936563 -1.61939366 -1.61941975 -1.61944406
#> hp          -0.07158548 -0.07158535 -0.07158524 -0.07158515 -0.07158506
#> qsec         1.44637168  1.44641273  1.44645059  1.44648557  1.44651791
#> vs          -2.43281406 -2.43275695 -2.43270339 -2.43265327 -2.43260645
#> wt          -2.29897881 -2.29913674 -2.29928357 -2.29942005 -2.29954689
#>                  0.0117      0.0109      0.0102      0.0095      0.0089
#> (Intercept) -1.99463325 -1.99488090 -1.99510959 -1.99532111 -1.99551700
#> am           6.42758105  6.42764150  6.42769768  6.42774991  6.42779849
#> carb         0.99662047  0.99665975  0.99669631  0.99673032  0.99676197
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.02527494  0.02527580  0.02527661  0.02527736  0.02527806
#> drat         2.79326190  2.79321553  2.79317242  2.79313235  2.79309512
#> gear        -1.61946670 -1.61948780 -1.61950745 -1.61952575 -1.61954280
#> hp          -0.07158499 -0.07158492 -0.07158487 -0.07158482 -0.07158477
#> qsec         1.44654785  1.44657559  1.44660131  1.44662516  1.44664731
#> vs          -2.43256277 -2.43252207 -2.43248416 -2.43244888 -2.43241605
#> wt          -2.29966475 -2.29977427 -2.29987604 -2.29997060 -2.30005848
#>                  0.0083        0.0077      0.0072      0.0067      0.0062
#> (Intercept) -1.99569861 -1.9968008330 -2.74629268 -2.74906758 -2.75171616
#> am           6.42784368  6.4278857333  6.44196870  6.44191389  6.44186134
#> carb         0.99679144  0.9968188594  0.96408448  0.96386322  0.96365183
#> cyl          0.00000000  0.0001835084  0.06956422  0.06986139  0.07014501
#> disp         0.02527871  0.0252765201  0.02434720  0.02434230  0.02433762
#> drat         2.79306051  2.7930738905  2.84873839  2.84902312  2.84929567
#> gear        -1.61955867 -1.6195936029 -1.58905265 -1.58883347 -1.58862459
#> hp          -0.07158474 -0.0715849164 -0.07160413 -0.07160441 -0.07160468
#> qsec         1.44666787  1.4467016065  1.45192643  1.45189525  1.45186540
#> vs          -2.43238551 -2.4323734496 -2.43140417 -2.43139333 -2.43138413
#> wt          -2.30014016 -2.3001812562 -2.23619110 -2.23575504 -2.23533768
#>                  0.0058      0.0054      0.0051      0.0047
#> (Intercept) -2.75424498 -2.75666014 -2.75896735 -2.76117197
#> am           6.44181098  6.44176273  6.44171650  6.44167221
#> carb         0.96344985  0.96325681  0.96307229  0.96289588
#> cyl          0.07041577  0.07067432  0.07092126  0.07115719
#> disp         0.02433313  0.02432884  0.02432473  0.02432079
#> drat         2.84955657  2.84980635  2.85004549  2.85027446
#> gear        -1.58842545 -1.58823553 -1.58805435 -1.58788146
#> hp          -0.07160493 -0.07160518 -0.07160540 -0.07160562
#> qsec         1.45183681  1.45180945  1.45178327  1.45175820
#> vs          -2.43137637 -2.43136988 -2.43136450 -2.43136009
#> wt          -2.23493819 -2.23455578 -2.23418970 -2.23383924
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   5   7   7   7  60   9   1   1   1   1   1   1   1   1   1   5
#>  [37]   8  55  87  21   1   1  64   1   1   1   1   1   1   1   1   1   1   1
#>  [55]   1  44   1   1   1   1   1   1   1   1  51 182   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1  50   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.724057219 4.405671573 4.108744054 3.831828456 3.573576043 3.332728978
#>   [7] 3.108114199 2.898637705 2.703279225 2.521087253 2.351174410 2.192713124
#>  [13] 2.044931598 1.907110052 1.778577218 1.658707066 1.546915761 1.442658816
#>  [19] 1.345428440 1.254751066 1.170185044 1.091318489 1.017767275 0.949173167
#>  [25] 0.885202072 0.825542415 0.769903619 0.718014692 0.669622905 0.624492562
#>  [31] 0.582403854 0.543151783 0.506545171 0.472405722 0.440567157 0.410874405
#>  [37] 0.383182846 0.357357604 0.333272898 0.310811420 0.289863770 0.270327922
#>  [43] 0.252108725 0.235117441 0.219271312 0.204493160 0.190711005 0.177857722
#>  [49] 0.165870707 0.154691576 0.144265882 0.134542844 0.125475107 0.117018504
#>  [55] 0.109131849 0.101776727 0.094917317 0.088520207 0.082554242 0.076990362
#>  [61] 0.071801469 0.066962291 0.062449256 0.058240385 0.054315178 0.050654517
#>  [67] 0.047240572 0.044056716 0.041087441 0.038318285 0.035735760 0.033327290
#>  [73] 0.031081142 0.028986377 0.027032792 0.025210873 0.023511744 0.021927131
#>  [79] 0.020449316 0.019071101 0.017785772 0.016587071 0.015469158 0.014426588
#>  [85] 0.013454284 0.012547511 0.011701850 0.010913185 0.010177673 0.009491732
#>  [91] 0.008852021 0.008255424 0.007699036 0.007180147 0.006696229 0.006244926
#>  [97] 0.005824039 0.005431518 0.005065452 0.004724057
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 24
#> 
#> $loss
#>   [1] 693.14667 603.17984 527.90920 465.21987 413.28594 370.53195 335.59940
#>   [8] 307.31772 284.67904 266.81630 252.98428 242.54313 234.94407 229.71702
#>  [15] 226.45988 224.82918 224.49562 224.49562 224.49562 224.49562 223.96083
#>  [22] 215.16530 207.99426 202.20407 162.01598 161.99281 161.99281 161.99281
#>  [29] 161.99281 161.99281 161.99281 161.99281 161.99281 161.99281 161.99281
#>  [36] 161.75263 160.63664 112.09163 106.04314 104.94674 104.94674 104.94674
#>  [43]  87.59113  87.59112  87.59112  87.59112  87.59112  87.59112  87.59112
#>  [50]  87.59112  87.59112  87.59112  87.59112  87.59112  87.59112  85.00092
#>  [57]  85.00091  85.00090  85.00090  85.00090  85.00090  85.00089  85.00089
#>  [64]  85.00089  84.60172  80.47558  80.47557  80.47555  80.47554  80.47554
#>  [71]  80.47553  80.47552  80.47552  80.47551  80.47551  80.47551  80.47551
#>  [78]  80.47550  80.47550  80.47550  80.47550  80.47550  80.47550  80.47549
#>  [85]  80.47549  80.47549  80.47549  80.47549  80.47549  80.47549  80.47549
#>  [92]  80.47549  80.47538  80.45310  80.45308  80.45306  80.45305  80.45303
#>  [99]  80.45302  80.45301
#> 
#> $linear.predictors
#>      4.7241   4.4057   4.1087   3.8318   3.5736   3.3327   3.1081   2.8986
#> 1  20.76667 21.27290 21.74501 22.18531 22.59593 22.97887 23.33601 23.66908
#> 2  20.76667 20.20981 19.69049 19.20616 18.75448 18.33324 17.94039 17.57401
#> 3  20.76667 20.74136 20.71775 20.69573 20.67520 20.65606 20.63820 20.62155
#> 4  20.76667 21.27290 21.74501 22.18531 22.59593 22.97887 23.33601 23.66908
#> 5  20.76667 21.27290 21.74501 22.18531 22.59593 22.97887 23.33601 23.66908
#> 6  20.76667 20.74136 20.71775 20.69573 20.67520 20.65606 20.63820 20.62155
#> 7  20.76667 20.74136 20.71775 20.69573 20.67520 20.65606 20.63820 20.62155
#> 8  20.76667 20.20981 19.69049 19.20616 18.75448 18.33324 17.94039 17.57401
#> 9  20.76667 20.20981 19.69049 19.20616 18.75448 18.33324 17.94039 17.57401
#> 10 20.76667 20.20981 19.69049 19.20616 18.75448 18.33324 17.94039 17.57401
#> 11 20.76667 21.27290 21.74501 22.18531 22.59593 22.97887 23.33601 23.66908
#> 12 20.76667 21.27290 21.74501 22.18531 22.59593 22.97887 23.33601 23.66908
#> 13 20.76667 21.27290 21.74501 22.18531 22.59593 22.97887 23.33601 23.66908
#> 14 20.76667 21.27290 21.74501 22.18531 22.59593 22.97887 23.33601 23.66908
#> 15 20.76667 20.20981 19.69049 19.20616 18.75448 18.33324 17.94039 17.57401
#> 16 20.76667 20.20981 19.69049 19.20616 18.75448 18.33324 17.94039 17.57401
#> 17 20.76667 20.20981 19.69049 19.20616 18.75448 18.33324 17.94039 17.57401
#> 18 20.76667 21.27290 21.74501 22.18531 22.59593 22.97887 23.33601 23.66908
#> 19 20.76667 20.20981 19.69049 19.20616 18.75448 18.33324 17.94039 17.57401
#> 20 20.76667 20.74136 20.71775 20.69573 20.67520 20.65606 20.63820 20.62155
#> 21 20.76667 21.27290 21.74501 22.18531 22.59593 22.97887 23.33601 23.66908
#>      2.7033   2.5211   2.3512   2.1927   2.0449   1.9071   1.7786   1.6587
#> 1  23.97970 24.26938 24.53954 24.79150 25.02647 25.24560 25.44997 25.64056
#> 2  17.23233 16.91368 16.61650 16.33935 16.08089 15.83984 15.61503 15.40538
#> 3  20.60602 20.59153 20.57802 20.56543 20.55368 20.54272 20.53250 20.52297
#> 4  23.97970 24.26938 24.53954 24.79150 25.02647 25.24560 25.44997 25.64056
#> 5  23.97970 24.26938 24.53954 24.79150 25.02647 25.24560 25.44997 25.64056
#> 6  20.60602 20.59153 20.57802 20.56543 20.55368 20.54272 20.53250 20.52297
#> 7  20.60602 20.59153 20.57802 20.56543 20.55368 20.54272 20.53250 20.52297
#> 8  17.23233 16.91368 16.61650 16.33935 16.08089 15.83984 15.61503 15.40538
#> 9  17.23233 16.91368 16.61650 16.33935 16.08089 15.83984 15.61503 15.40538
#> 10 17.23233 16.91368 16.61650 16.33935 16.08089 15.83984 15.61503 15.40538
#> 11 23.97970 24.26938 24.53954 24.79150 25.02647 25.24560 25.44997 25.64056
#> 12 23.97970 24.26938 24.53954 24.79150 25.02647 25.24560 25.44997 25.64056
#> 13 23.97970 24.26938 24.53954 24.79150 25.02647 25.24560 25.44997 25.64056
#> 14 23.97970 24.26938 24.53954 24.79150 25.02647 25.24560 25.44997 25.64056
#> 15 17.23233 16.91368 16.61650 16.33935 16.08089 15.83984 15.61503 15.40538
#> 16 17.23233 16.91368 16.61650 16.33935 16.08089 15.83984 15.61503 15.40538
#> 17 17.23233 16.91368 16.61650 16.33935 16.08089 15.83984 15.61503 15.40538
#> 18 23.97970 24.26938 24.53954 24.79150 25.02647 25.24560 25.44997 25.64056
#> 19 17.23233 16.91368 16.61650 16.33935 16.08089 15.83984 15.61503 15.40538
#> 20 20.60602 20.59153 20.57802 20.56543 20.55368 20.54272 20.53250 20.52297
#> 21 23.97970 24.26938 24.53954 24.79150 25.02647 25.24560 25.44997 25.64056
#>      1.5469   1.4427   1.3454   1.2548   1.1702   1.0913   1.0178   0.9492
#> 1  25.77416 25.77416 25.77416 25.77416 25.78221 25.92376 26.05578 26.17890
#> 2  15.25843 15.25843 15.25843 15.25843 15.25629 15.21869 15.18362 15.15092
#> 3  20.51629 20.51629 20.51629 20.51629 20.50805 20.36318 20.22807 20.10206
#> 4  25.77416 25.77416 25.77416 25.77416 25.75981 25.50767 25.27252 25.05321
#> 5  25.77416 25.77416 25.77416 25.77416 25.75981 25.50767 25.27252 25.05321
#> 6  20.51629 20.51629 20.51629 20.51629 20.50805 20.36318 20.22807 20.10206
#> 7  20.51629 20.51629 20.51629 20.51629 20.50805 20.36318 20.22807 20.10206
#> 8  15.25843 15.25843 15.25843 15.25843 15.25629 15.21869 15.18362 15.15092
#> 9  15.25843 15.25843 15.25843 15.25843 15.25629 15.21869 15.18362 15.15092
#> 10 15.25843 15.25843 15.25843 15.25843 15.25629 15.21869 15.18362 15.15092
#> 11 25.77416 25.77416 25.77416 25.77416 25.78221 25.92376 26.05578 26.17890
#> 12 25.77416 25.77416 25.77416 25.77416 25.78221 25.92376 26.05578 26.17890
#> 13 25.77416 25.77416 25.77416 25.77416 25.78221 25.92376 26.05578 26.17890
#> 14 25.77416 25.77416 25.77416 25.77416 25.75981 25.50767 25.27252 25.05321
#> 15 15.25843 15.25843 15.25843 15.25843 15.25629 15.21869 15.18362 15.15092
#> 16 15.25843 15.25843 15.25843 15.25843 15.25629 15.21869 15.18362 15.15092
#> 17 15.25843 15.25843 15.25843 15.25843 15.25629 15.21869 15.18362 15.15092
#> 18 25.77416 25.77416 25.77416 25.77416 25.78221 25.92376 26.05578 26.17890
#> 19 15.25843 15.25843 15.25843 15.25843 15.27868 15.63478 15.96688 16.27661
#> 20 20.51629 20.51629 20.51629 20.51629 20.53045 20.77927 21.01133 21.22775
#> 21 25.77416 25.77416 25.77416 25.77416 25.78221 25.92376 26.05578 26.17890
#>      0.8852   0.8255   0.7699   0.7180   0.6696   0.6245   0.5824   0.5432
#> 1  26.16540 26.17282 26.17276 26.17272 26.17270 26.17269 26.17268 26.17268
#> 2  16.46173 16.48541 16.48558 16.48567 16.48572 16.48575 16.48577 16.48578
#> 3  19.47192 19.45157 19.45151 19.45148 19.45146 19.45145 19.45145 19.45144
#> 4  23.40293 23.35650 23.35628 23.35615 23.35608 23.35604 23.35601 23.35600
#> 5  23.52994 23.48598 23.48577 23.48565 23.48558 23.48554 23.48551 23.48550
#> 6  19.53542 19.51631 19.51626 19.51623 19.51621 19.51620 19.51620 19.51619
#> 7  19.53542 19.51631 19.51626 19.51623 19.51621 19.51620 19.51620 19.51619
#> 8  15.54090 15.54664 15.54675 15.54681 15.54685 15.54687 15.54688 15.54689
#> 9  15.38214 15.38478 15.38488 15.38494 15.38497 15.38499 15.38500 15.38501
#> 10 10.41288 10.31863 10.31845 10.31835 10.31829 10.31826 10.31824 10.31823
#> 11 26.54643 26.56128 26.56124 26.56121 26.56120 26.56119 26.56119 26.56118
#> 12 28.40395 28.45502 28.45508 28.45511 28.45513 28.45515 28.45515 28.45516
#> 13 27.70539 27.74284 27.74287 27.74288 27.74288 27.74289 27.74289 27.74289
#> 14 25.70499 25.70344 25.70334 25.70329 25.70326 25.70324 25.70323 25.70323
#> 15 16.20771 16.22644 16.22659 16.22667 16.22672 16.22675 16.22677 16.22678
#> 16 16.47760 16.50160 16.50176 16.50186 16.50191 16.50194 16.50196 16.50197
#> 17 15.19163 15.19055 15.19064 15.19069 15.19072 15.19074 15.19075 15.19076
#> 18 26.73694 26.75551 26.75548 26.75546 26.75545 26.75544 26.75544 26.75544
#> 19 17.31904 17.35944 17.35966 17.35978 17.35985 17.35989 17.35991 17.35992
#> 20 21.66284 21.68520 21.68527 21.68531 21.68534 21.68535 21.68536 21.68536
#> 21 24.70479 24.68373 24.68358 24.68350 24.68345 24.68342 24.68341 24.68340
#>      0.5065   0.4724   0.4406   0.4109   0.3832   0.3574   0.3333   0.3108
#> 1  26.17267 26.17267 26.17267 26.18056 26.22141 25.96972 26.11517 26.20939
#> 2  16.48579 16.48579 16.48579 16.49853 16.55933 15.80612 15.98309 16.09647
#> 3  19.45144 19.45144 19.45144 19.46873 19.55655 19.35666 19.29523 19.25397
#> 4  23.35599 23.35599 23.35598 23.35606 23.36069 23.81012 23.28504 22.94571
#> 5  23.48549 23.48549 23.48549 23.48544 23.48932 22.22595 21.83371 21.58052
#> 6  19.51619 19.51619 19.51619 19.50210 19.43230 20.50264 20.13662 19.90051
#> 7  19.51619 19.51619 19.51619 19.50210 19.43230 20.50264 20.13662 19.90051
#> 8  15.54689 15.54690 15.54690 15.55007 15.56385 15.34295 15.48915 15.58262
#> 9  15.38501 15.38502 15.38502 15.38834 15.40306 15.34295 15.48915 15.58262
#> 10 10.31822 10.31822 10.31822 10.31583 10.30725 12.57648 12.84041 13.01031
#> 11 26.56118 26.56118 26.56118 26.56871 26.60733 28.07836 28.30046 28.44421
#> 12 28.45516 28.45516 28.45516 28.45048 28.42579 30.59153 31.17719 31.55636
#> 13 27.74289 27.74289 27.74289 27.74932 27.78114 28.41193 28.69512 28.87844
#> 14 25.70323 25.70322 25.70322 25.71155 25.75510 21.67541 22.02305 22.24670
#> 15 16.22678 16.22679 16.22679 16.23977 16.30206 16.58705 16.56777 16.55396
#> 16 16.50198 16.50198 16.50198 16.51470 16.57541 17.34480 17.49835 17.59648
#> 17 15.19076 15.19076 15.19076 15.18383 15.14724 12.62475 13.12431 13.44687
#> 18 26.75544 26.75543 26.75543 26.75234 26.73743 27.21973 26.78482 26.50553
#> 19 17.35993 17.35994 17.35994 17.35098 17.30192 16.16409 15.79815 15.56415
#> 20 21.68536 21.68536 21.68537 21.64836 21.46127 20.47599 19.76081 19.30029
#> 21 24.68339 24.68339 24.68339 24.68223 24.67922 25.49014 25.76579 25.94439
#>      0.2899   0.2703   0.2521   0.2351   0.2193   0.2045   0.1907   0.1779
#> 1  26.20913 26.20898 25.50552 25.50556 25.50558 25.50560 25.50562 25.50563
#> 2  16.09645 16.09644 16.16797 16.16785 16.16777 16.16770 16.16766 16.16763
#> 3  19.25390 19.25391 18.54340 18.54334 18.54330 18.54327 18.54325 18.54323
#> 4  22.94599 22.94619 22.41150 22.41125 22.41107 22.41093 22.41084 22.41077
#> 5  21.58083 21.58102 24.27812 24.27851 24.27878 24.27897 24.27910 24.27919
#> 6  19.90082 19.90100 18.23333 18.23320 18.23311 18.23304 18.23300 18.23297
#> 7  19.90082 19.90100 18.82525 18.82520 18.82517 18.82515 18.82514 18.82513
#> 8  15.58259 15.58258 16.19923 16.19922 16.19922 16.19921 16.19920 16.19920
#> 9  15.58259 15.58258 16.59384 16.59389 16.59393 16.59395 16.59396 16.59397
#> 10 13.01029 13.01027 13.87466 13.87493 13.87512 13.87526 13.87536 13.87543
#> 11 28.44399 28.44385 28.58847 28.58850 28.58851 28.58852 28.58853 28.58853
#> 12 31.55627 31.55614 31.29307 31.29294 31.29284 31.29277 31.29272 31.29269
#> 13 28.87825 28.87811 29.54223 29.54232 29.54238 29.54243 29.54246 29.54248
#> 14 22.24677 22.24679 21.96097 21.96091 21.96086 21.96083 21.96080 21.96079
#> 15 16.55389 16.55388 16.04007 16.03976 16.03953 16.03937 16.03925 16.03916
#> 16 17.59646 17.59646 17.78954 17.78932 17.78916 17.78904 17.78896 17.78890
#> 17 13.44692 13.44690 12.78361 12.78370 12.78377 12.78382 12.78387 12.78390
#> 18 26.50557 26.50559 27.53383 27.53362 27.53345 27.53333 27.53325 27.53318
#> 19 15.56412 15.56409 15.34008 15.34052 15.34084 15.34108 15.34125 15.34138
#> 20 19.30020 19.30018 19.07717 19.07719 19.07722 19.07723 19.07724 19.07725
#> 21 25.94417 25.94402 25.51814 25.51829 25.51840 25.51848 25.51853 25.51858
#>      0.1659   0.1547   0.1443   0.1345   0.1255   0.1170   0.1091   0.1018
#> 1  25.50564 25.50565 25.50565 25.50566 25.50566 25.50566 25.50566 25.22408
#> 2  16.16761 16.16759 16.16758 16.16757 16.16756 16.16756 16.16756 16.30282
#> 3  18.54322 18.54321 18.54321 18.54320 18.54320 18.54320 18.54320 18.62972
#> 4  22.41072 22.41068 22.41066 22.41064 22.41063 22.41062 22.41061 22.08322
#> 5  24.27926 24.27930 24.27933 24.27935 24.27936 24.27937 24.27937 24.30101
#> 6  18.23295 18.23294 18.23293 18.23293 18.23292 18.23292 18.23292 18.52509
#> 7  18.82513 18.82513 18.82512 18.82512 18.82512 18.82512 18.82512 19.22654
#> 8  16.19919 16.19919 16.19919 16.19919 16.19919 16.19918 16.19918 16.36785
#> 9  16.59398 16.59398 16.59398 16.59399 16.59399 16.59399 16.59399 16.83548
#> 10 13.87548 13.87551 13.87554 13.87556 13.87557 13.87558 13.87558 13.66629
#> 11 28.58854 28.58854 28.58854 28.58854 28.58854 28.58854 28.58854 28.77415
#> 12 31.29267 31.29265 31.29264 31.29263 31.29263 31.29262 31.29262 31.75315
#> 13 29.54250 29.54251 29.54251 29.54252 29.54252 29.54252 29.54252 29.86954
#> 14 21.96078 21.96077 21.96076 21.96076 21.96076 21.96075 21.96075 21.10673
#> 15 16.03910 16.03906 16.03903 16.03901 16.03900 16.03899 16.03898 16.19783
#> 16 17.78885 17.78882 17.78880 17.78878 17.78877 17.78877 17.78876 18.17900
#> 17 12.78393 12.78394 12.78396 12.78397 12.78398 12.78398 12.78399 12.28259
#> 18 27.53313 27.53310 27.53308 27.53306 27.53305 27.53304 27.53303 27.10737
#> 19 15.34147 15.34153 15.34158 15.34161 15.34163 15.34165 15.34166 15.69448
#> 20 19.07726 19.07726 19.07727 19.07727 19.07727 19.07727 19.07727 18.76630
#> 21 25.51861 25.51863 25.51864 25.51866 25.51866 25.51867 25.51867 25.20677
#>      0.0949   0.0885   0.0826   0.0770   0.0718   0.0670   0.0624   0.0582
#> 1  25.22447 25.22480 25.22508 25.22532 25.22551 25.22568 25.22582 25.22593
#> 2  16.30276 16.30270 16.30264 16.30259 16.30254 16.30250 16.30246 16.30243
#> 3  18.63006 18.63032 18.63052 18.63066 18.63076 18.63083 18.63086 18.63088
#> 4  22.08272 22.08233 22.08204 22.08183 22.08169 22.08159 22.08153 22.08150
#> 5  24.30070 24.30044 24.30022 24.30004 24.29989 24.29977 24.29967 24.29959
#> 6  18.52496 18.52484 18.52472 18.52462 18.52453 18.52445 18.52437 18.52431
#> 7  19.22643 19.22632 19.22620 19.22609 19.22598 19.22589 19.22579 19.22571
#> 8  16.36786 16.36785 16.36784 16.36781 16.36778 16.36775 16.36771 16.36768
#> 9  16.83551 16.83551 16.83549 16.83546 16.83542 16.83537 16.83532 16.83527
#> 10 13.66648 13.66665 13.66680 13.66693 13.66704 13.66713 13.66721 13.66728
#> 11 28.77436 28.77451 28.77462 28.77470 28.77475 28.77477 28.77478 28.77478
#> 12 31.75296 31.75279 31.75263 31.75249 31.75236 31.75226 31.75216 31.75208
#> 13 29.86971 29.86982 29.86989 29.86993 29.86994 29.86993 29.86991 29.86987
#> 14 21.10649 21.10636 21.10631 21.10633 21.10638 21.10647 21.10658 21.10671
#> 15 16.19778 16.19772 16.19766 16.19760 16.19754 16.19749 16.19744 16.19739
#> 16 18.17883 18.17866 18.17850 18.17835 18.17822 18.17809 18.17798 18.17788
#> 17 12.28261 12.28267 12.28276 12.28286 12.28298 12.28311 12.28323 12.28336
#> 18 27.10667 27.10615 27.10575 27.10547 27.10528 27.10515 27.10508 27.10505
#> 19 15.69505 15.69548 15.69581 15.69605 15.69622 15.69634 15.69641 15.69645
#> 20 18.76646 18.76661 18.76675 18.76687 18.76697 18.76707 18.76715 18.76722
#> 21 25.20715 25.20747 25.20776 25.20800 25.20820 25.20838 25.20852 25.20864
#>      0.0543   0.0507   0.0472   0.0441   0.0411   0.0383   0.0357   0.0333
#> 1  25.22907 24.93117 24.93108 24.93100 24.93093 24.93088 24.93083 24.93079
#> 2  16.40987 17.21833 17.21842 17.21851 17.21861 17.21872 17.21882 17.21892
#> 3  18.66011 18.88252 18.88250 18.88249 18.88251 18.88254 18.88257 18.88262
#> 4  21.95413 22.26278 22.26246 22.26217 22.26192 22.26170 22.26150 22.26132
#> 5  24.39599 24.67543 24.67580 24.67612 24.67638 24.67661 24.67680 24.67696
#> 6  18.45450 18.02520 18.02496 18.02474 18.02455 18.02438 18.02424 18.02410
#> 7  19.17784 18.89199 18.89189 18.89179 18.89171 18.89163 18.89157 18.89151
#> 8  16.40771 15.67932 15.67933 15.67933 15.67933 15.67932 15.67931 15.67930
#> 9  16.86932 16.14247 16.14255 16.14261 16.14265 16.14269 16.14271 16.14273
#> 10 13.21384 13.71898 13.71912 13.71922 13.71930 13.71935 13.71938 13.71940
#> 11 28.71912 28.28547 28.28547 28.28547 28.28546 28.28545 28.28544 28.28542
#> 12 31.73103 32.55198 32.55217 32.55234 32.55249 32.55261 32.55272 32.55282
#> 13 29.95738 30.01520 30.01543 30.01564 30.01581 30.01597 30.01611 30.01623
#> 14 21.36942 21.41521 21.41512 21.41505 21.41499 21.41494 21.41489 21.41485
#> 15 16.21901 16.45818 16.45803 16.45792 16.45783 16.45777 16.45772 16.45769
#> 16 18.19721 18.01119 18.01111 18.01104 18.01098 18.01093 18.01090 18.01087
#> 17 12.36922 12.32089 12.32088 12.32088 12.32087 12.32086 12.32085 12.32085
#> 18 27.02386 26.46951 26.46914 26.46881 26.46849 26.46820 26.46793 26.46767
#> 19 15.85170 15.76533 15.76553 15.76572 15.76590 15.76608 15.76624 15.76640
#> 20 18.83905 19.61381 19.61401 19.61420 19.61436 19.61451 19.61464 19.61477
#> 21 25.05063 24.76502 24.76499 24.76496 24.76492 24.76488 24.76484 24.76479
#>      0.0311   0.0290   0.0270   0.0252   0.0235   0.0219   0.0204   0.0191
#> 1  24.93075 24.93072 24.93070 24.93067 24.93065 24.93063 24.93061 24.93060
#> 2  17.21902 17.21912 17.21921 17.21930 17.21938 17.21946 17.21954 17.21961
#> 3  18.88266 18.88271 18.88276 18.88281 18.88286 18.88290 18.88295 18.88299
#> 4  22.26116 22.26101 22.26088 22.26076 22.26064 22.26054 22.26045 22.26036
#> 5  24.67710 24.67721 24.67732 24.67740 24.67748 24.67755 24.67761 24.67766
#> 6  18.02399 18.02388 18.02378 18.02370 18.02362 18.02355 18.02348 18.02342
#> 7  18.89145 18.89140 18.89136 18.89132 18.89128 18.89125 18.89121 18.89119
#> 8  15.67928 15.67927 15.67926 15.67925 15.67924 15.67923 15.67922 15.67921
#> 9  16.14275 16.14276 16.14277 16.14278 16.14278 16.14279 16.14279 16.14279
#> 10 13.71941 13.71941 13.71941 13.71940 13.71939 13.71938 13.71936 13.71935
#> 11 28.28540 28.28539 28.28537 28.28535 28.28533 28.28532 28.28530 28.28528
#> 12 32.55290 32.55298 32.55304 32.55311 32.55316 32.55322 32.55326 32.55331
#> 13 30.01634 30.01644 30.01652 30.01660 30.01668 30.01674 30.01680 30.01686
#> 14 21.41482 21.41479 21.41477 21.41475 21.41473 21.41471 21.41470 21.41468
#> 15 16.45767 16.45766 16.45765 16.45765 16.45765 16.45766 16.45767 16.45767
#> 16 18.01084 18.01082 18.01081 18.01080 18.01079 18.01078 18.01077 18.01076
#> 17 12.32084 12.32083 12.32083 12.32083 12.32082 12.32082 12.32082 12.32082
#> 18 26.46744 26.46722 26.46701 26.46682 26.46664 26.46648 26.46632 26.46618
#> 19 15.76654 15.76668 15.76680 15.76692 15.76702 15.76712 15.76721 15.76730
#> 20 19.61488 19.61499 19.61509 19.61518 19.61527 19.61535 19.61542 19.61549
#> 21 24.76475 24.76471 24.76466 24.76462 24.76458 24.76454 24.76451 24.76447
#>      0.0178   0.0166   0.0155   0.0144   0.0135   0.0125   0.0117   0.0109
#> 1  24.93058 24.93057 24.93056 24.93055 24.93054 24.93053 24.93052 24.93051
#> 2  17.21967 17.21973 17.21979 17.21984 17.21989 17.21994 17.21998 17.22002
#> 3  18.88303 18.88307 18.88310 18.88314 18.88317 18.88319 18.88322 18.88325
#> 4  22.26028 22.26020 22.26013 22.26007 22.26001 22.25996 22.25990 22.25986
#> 5  24.67770 24.67775 24.67778 24.67782 24.67785 24.67788 24.67790 24.67793
#> 6  18.02336 18.02331 18.02327 18.02322 18.02318 18.02314 18.02311 18.02308
#> 7  18.89116 18.89114 18.89111 18.89109 18.89107 18.89106 18.89104 18.89102
#> 8  15.67920 15.67920 15.67919 15.67918 15.67918 15.67917 15.67916 15.67916
#> 9  16.14279 16.14280 16.14280 16.14280 16.14280 16.14280 16.14280 16.14280
#> 10 13.71934 13.71932 13.71931 13.71930 13.71929 13.71927 13.71926 13.71925
#> 11 28.28527 28.28526 28.28524 28.28523 28.28522 28.28521 28.28520 28.28519
#> 12 32.55335 32.55339 32.55342 32.55346 32.55349 32.55351 32.55354 32.55357
#> 13 30.01691 30.01696 30.01700 30.01704 30.01708 30.01711 30.01714 30.01717
#> 14 21.41467 21.41466 21.41465 21.41464 21.41463 21.41463 21.41462 21.41461
#> 15 16.45768 16.45769 16.45770 16.45771 16.45772 16.45773 16.45773 16.45774
#> 16 18.01076 18.01076 18.01075 18.01075 18.01075 18.01075 18.01074 18.01074
#> 17 12.32082 12.32081 12.32081 12.32081 12.32081 12.32081 12.32081 12.32081
#> 18 26.46604 26.46592 26.46580 26.46569 26.46559 26.46550 26.46541 26.46533
#> 19 15.76737 15.76745 15.76751 15.76758 15.76763 15.76768 15.76773 15.76778
#> 20 19.61556 19.61562 19.61567 19.61572 19.61577 19.61582 19.61586 19.61590
#> 21 24.76444 24.76441 24.76438 24.76436 24.76433 24.76431 24.76429 24.76427
#>      0.0102   0.0095   0.0089   0.0083   0.0077   0.0072   0.0067   0.0062
#> 1  24.93050 24.93050 24.93049 24.93048 24.93039 24.92075 24.92074 24.92074
#> 2  17.22006 17.22010 17.22013 17.22016 17.22013 17.19832 17.19819 17.19807
#> 3  18.88327 18.88329 18.88331 18.88333 18.88333 18.87955 18.87954 18.87953
#> 4  22.25981 22.25977 22.25974 22.25970 22.25952 22.22172 22.22166 22.22160
#> 5  24.67795 24.67797 24.67799 24.67801 24.67794 24.67038 24.67030 24.67022
#> 6  18.02305 18.02302 18.02300 18.02297 18.02309 18.05791 18.05811 18.05830
#> 7  18.89101 18.89100 18.89098 18.89097 18.89111 18.92907 18.92925 18.92942
#> 8  15.67915 15.67915 15.67915 15.67914 15.67933 15.72007 15.72022 15.72036
#> 9  16.14280 16.14280 16.14280 16.14280 16.14300 16.18903 16.18919 16.18934
#> 10 13.71925 13.71924 13.71923 13.71922 13.71900 13.68476 13.68462 13.68449
#> 11 28.28518 28.28517 28.28517 28.28516 28.28517 28.31289 28.31303 28.31316
#> 12 32.55359 32.55361 32.55363 32.55365 32.55370 32.55666 32.55661 32.55656
#> 13 30.01720 30.01723 30.01725 30.01727 30.01733 30.03882 30.03886 30.03890
#> 14 21.41460 21.41460 21.41459 21.41459 21.41450 21.35715 21.35690 21.35666
#> 15 16.45775 16.45776 16.45776 16.45777 16.45782 16.45815 16.45816 16.45817
#> 16 18.01074 18.01074 18.01074 18.01074 18.01085 18.04270 18.04284 18.04297
#> 17 12.32081 12.32081 12.32081 12.32081 12.32078 12.29092 12.29076 12.29062
#> 18 26.46526 26.46519 26.46512 26.46506 26.46487 26.45146 26.45154 26.45160
#> 19 15.76782 15.76786 15.76790 15.76793 15.76787 15.79153 15.79162 15.79171
#> 20 19.61593 19.61597 19.61600 19.61603 19.61617 19.57497 19.57469 19.57442
#> 21 24.76425 24.76424 24.76422 24.76421 24.76409 24.75318 24.75317 24.75315
#>      0.0058   0.0054   0.0051   0.0047
#> 1  24.92074 24.92074 24.92073 24.92073
#> 2  17.19795 17.19783 17.19772 17.19762
#> 3  18.87951 18.87950 18.87949 18.87948
#> 4  22.22155 22.22150 22.22145 22.22140
#> 5  24.67015 24.67007 24.67001 24.66994
#> 6  18.05848 18.05866 18.05883 18.05899
#> 7  18.92959 18.92975 18.92990 18.93004
#> 8  15.72049 15.72062 15.72074 15.72086
#> 9  16.18948 16.18961 16.18974 16.18987
#> 10 13.68437 13.68425 13.68413 13.68402
#> 11 28.31328 28.31340 28.31351 28.31362
#> 12 32.55652 32.55647 32.55643 32.55639
#> 13 30.03894 30.03897 30.03900 30.03903
#> 14 21.35644 21.35622 21.35601 21.35581
#> 15 16.45818 16.45819 16.45820 16.45821
#> 16 18.04310 18.04323 18.04334 18.04345
#> 17 12.29048 12.29035 12.29022 12.29010
#> 18 26.45167 26.45173 26.45180 26.45186
#> 19 15.79180 15.79188 15.79196 15.79203
#> 20 19.57416 19.57391 19.57368 19.57345
#> 21 24.75314 24.75312 24.75311 24.75310
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 22.8 18.7 18.1 24.4 22.8 19.2 17.8 17.3 15.2 14.7 32.4 30.4 33.9 21.5 15.5
#> [16] 15.2 13.3 26.0 15.8 19.7 21.4
#> 
#> $X
#>               am       carb         cyl       disp         drat       gear
#>  [1,]  1.2747549 -1.1443177 -1.05999788 -0.9143587  0.285050273  0.4082483
#>  [2,] -0.7844645 -0.3933592  1.16599767  1.3962069 -0.977069912 -1.0206207
#>  [3,] -0.7844645 -1.1443177  0.05299989  0.1584039 -1.680251158 -1.0206207
#>  [4,] -0.7844645 -0.3933592 -1.05999788 -0.5595219 -0.003434341  0.4082483
#>  [5,] -0.7844645 -0.3933592 -1.05999788 -0.6136185  0.411262292  0.4082483
#>  [6,] -0.7844645  1.1085577  0.05299989 -0.3678916  0.411262292  0.4082483
#>  [7,] -0.7844645  1.1085577  0.05299989 -0.3678916  0.411262292  0.4082483
#>  [8,] -0.7844645  0.3575993  1.16599767  0.6241845 -1.121312219 -1.0206207
#>  [9,] -0.7844645  0.3575993  1.16599767  0.6241845 -1.121312219 -1.0206207
#> [10,] -0.7844645  1.1085577  1.16599767  2.1297197 -0.832827605 -1.0206207
#> [11,]  1.2747549 -1.1443177 -1.05999788 -1.1830078  0.699746905  0.4082483
#> [12,]  1.2747549 -0.3933592 -1.05999788 -1.2105146  2.232321416  0.4082483
#> [13,]  1.2747549 -1.1443177 -1.05999788 -1.2526916  0.952170942  0.4082483
#> [14,] -0.7844645 -1.1443177 -1.05999788 -0.8034149  0.014595948 -1.0206207
#> [15,] -0.7844645 -0.3933592  1.16599767  1.0111126 -1.680251158 -1.0206207
#> [16,] -0.7844645 -0.3933592  1.16599767  0.8827478 -0.977069912 -1.0206207
#> [17,] -0.7844645  1.1085577  1.16599767  1.3045177  0.068686813 -1.0206207
#> [18,]  1.2747549 -0.3933592 -1.05999788 -0.8015811  1.330806998  1.8371173
#> [19,]  1.2747549  1.1085577  1.16599767  1.3136867  0.952170942  1.8371173
#> [20,]  1.2747549  2.6104747  0.05299989 -0.5751090 -0.129646359  1.8371173
#> [21,]  1.2747549 -0.3933592 -1.05999788 -0.7951629  0.753837770  0.4082483
#>               hp        qsec         vs          wt
#>  [1,] -0.6944179  0.22966645  0.9534626 -0.95369015
#>  [2,]  0.6690799 -0.61675328 -1.0488088  0.41864343
#>  [3,] -0.4948816  1.08673297  0.9534626  0.44314939
#>  [4,] -1.2098865  0.96961829  0.9534626  0.11231897
#>  [5,] -0.6611618  2.51340271  0.9534626  0.06330706
#>  [6,] -0.1955772  0.06464122  0.9534626  0.41864343
#>  [7,] -0.1955772  0.38404489  0.9534626  0.41864343
#>  [8,]  0.7522200 -0.30799640 -1.0488088  0.77397980
#>  [9,]  0.7522200 -0.09506062 -1.0488088  0.83524469
#> [10,]  1.5836211 -0.40381750 -1.0488088  2.75283581
#> [11,] -1.1433745  0.68747838  0.9534626 -1.10072589
#> [12,] -1.3761668  0.18175590  0.9534626 -1.81752513
#> [13,] -1.1600025  0.91638434  0.9534626 -1.54795960
#> [14,] -0.6279058  0.97494168  0.9534626 -0.77602196
#> [15,]  0.2533794 -0.69660420 -1.0488088  0.51666726
#> [16,]  0.2533794 -0.46769823 -1.0488088  0.41251694
#> [17,]  1.8330414 -1.47381980 -1.0488088  0.90876256
#> [18,] -0.7276739 -0.78710191 -1.0488088 -1.17424376
#> [19,]  2.1489739 -1.95824870 -1.0488088  0.08781301
#> [20,]  0.6690799 -1.42590925 -1.0488088 -0.40230612
#> [21,] -0.4283695  0.22434306  0.9534626 -0.39005315
#> attr(,"center")
#>  [1]   0.3809524   2.5238095   5.9047619 207.7238095   3.6919048   3.7142857
#>  [7] 134.7619048  18.1785714   0.5238095   3.0983333
#> attr(,"scale")
#>  [1]   0.4856209   1.3316316   1.7969488 109.0642052   0.5546223   0.6998542
#>  [7]  60.1394449   1.8785006   0.4994328   0.8161281
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 50
#> 
#> $lambda.min
#> [1] 0.1546916
#> 
#> $null.dev
#> [1] 33.00698
#> 
#> $Bias
#> [1] 7.57987
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 21.23221