Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 25.358656 23.705624 20.705830 17.862878 15.603585 13.837547 12.487109
#>   [8] 11.485629 10.775994 10.309310 10.043790  9.943770  9.978871 10.123264
#>  [15] 10.355040 10.606427 10.722018 10.889062 10.919555 10.919555 10.919555
#>  [22] 10.873030 10.852394 10.913898 10.955648 11.011699 10.838843 10.828360
#>  [29] 10.512418 10.564968 10.566083  9.672466  9.680936  9.710595  9.747201
#>  [36]  9.772111  9.798286  9.836854  9.874793 10.263718 10.758706 13.169890
#>  [43] 12.972336 12.933561 12.664879 12.664678 12.373526 12.779514 12.933674
#>  [50] 13.014232 13.067830 13.070182 16.351250 16.341625 16.257926 21.712545
#>  [57] 21.819812 21.820886 32.820221 32.836272 32.747227 32.712351 33.004136
#>  [64] 33.006623 33.018850 33.021206 33.023502 32.763569 32.641879 32.643989
#>  [71] 32.608331 32.610376 33.532077 33.534800 33.537477 33.540106 33.542687
#>  [78] 33.545218 33.547700 33.550131 33.547386 33.549714 33.551994 33.553847
#>  [85] 33.556036 33.541136 33.542990 33.559383 33.561222 33.556761 34.070361
#>  [92] 34.072715 34.079473 34.081860 34.084224 34.084858 34.087129 34.089367
#>  [99] 34.075663 34.077701
#> 
#> $cvse
#>   [1]  8.161713  8.152860  7.642498  7.041268  6.509898  6.029536  5.589440
#>   [8]  5.185044  4.816365  4.486708  4.201554  3.967429  3.790622  3.675760
#>  [15]  3.624487  3.628278  3.676322  3.774558  3.793688  3.793688  3.793688
#>  [22]  3.766113  3.739920  3.711554  3.687263  3.656141  3.637080  3.621290
#>  [29]  3.624155  3.613630  3.622280  3.537958  3.528030  3.518139  3.509195
#>  [36]  3.502451  3.498072  3.499949  3.501740  3.521576  3.501919  3.997510
#>  [43]  3.928792  3.926535  3.867316  3.867231  3.623389  3.557197  3.558817
#>  [50]  3.440057  3.432812  3.428277  4.309780  4.311910  4.320709  7.153751
#>  [57]  7.189974  7.190970 10.363035 10.362261 10.367231 10.365409 10.328225
#>  [64] 10.329435 10.319116 10.320259 10.321368 10.377315 10.365363 10.366372
#>  [71] 10.373193 10.374127 10.774641 10.775911 10.777151 10.778363 10.779548
#>  [78] 10.780705 10.781835 10.782940 10.784809 10.785864 10.786895 10.787792
#>  [85] 10.788775 10.791856 10.792825 10.791754 10.792671 10.794451 11.036533
#>  [92] 11.037720 11.039266 11.040416 11.041544 11.043364 11.044455 11.045525
#>  [99] 11.048639 11.049689
#> 
#> $fold
#>  [1]  4  7  3  6  8  2  1  7  4  3  9  9  8 10  1 10  2  1  5  6  5
#> 
#> $lambda
#>   [1] 4.162506256 3.881967266 3.620335665 3.376337158 3.148783334 2.936565875
#>   [7] 2.738651164 2.554075241 2.381939118 2.221404394 2.071689172 1.932064256
#>  [13] 1.801849591 1.680410960 1.567156885 1.461535756 1.363033138 1.271169266
#>  [19] 1.185496712 1.105598201 1.031084583 0.961592933 0.896784788 0.836344494
#>  [25] 0.779977675 0.727409790 0.678384805 0.632663939 0.590024507 0.550258830
#>  [31] 0.513173227 0.478587069 0.446331902 0.416250626 0.388196727 0.362033566
#>  [37] 0.337633716 0.314878333 0.293656588 0.273865116 0.255407524 0.238193912
#>  [43] 0.222140439 0.207168917 0.193206426 0.180184959 0.168041096 0.156715689
#>  [49] 0.146153576 0.136303314 0.127116927 0.118549671 0.110559820 0.103108458
#>  [55] 0.096159293 0.089678479 0.083634449 0.077997767 0.072740979 0.067838480
#>  [61] 0.063266394 0.059002451 0.055025883 0.051317323 0.047858707 0.044633190
#>  [67] 0.041625063 0.038819673 0.036203357 0.033763372 0.031487833 0.029365659
#>  [73] 0.027386512 0.025540752 0.023819391 0.022214044 0.020716892 0.019320643
#>  [79] 0.018018496 0.016804110 0.015671569 0.014615358 0.013630331 0.012711693
#>  [85] 0.011854967 0.011055982 0.010310846 0.009615929 0.008967848 0.008363445
#>  [91] 0.007799777 0.007274098 0.006783848 0.006326639 0.005900245 0.005502588
#>  [97] 0.005131732 0.004785871 0.004463319 0.004162506
#> 
#> $fit
#> $beta
#>               4.1625     3.8820     3.6203     3.3763     3.1488    2.9366
#> (Intercept) 20.58571 22.0320043 23.3808193 24.6387286 25.8118591 26.905924
#> am           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> cyl          0.00000 -0.2410483 -0.4658508 -0.6755024 -0.8710241 -1.053368
#> disp         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> wt           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#>                2.7387    2.5541    2.3819    2.2214    2.0717    1.9321
#> (Intercept) 27.926253 28.877816 29.765246 30.592866 31.364708 32.084530
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.223423 -1.382017 -1.529922 -1.667859 -1.796499 -1.916469
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.8018    1.6804   1.5672    1.4615    1.3630    1.2712
#> (Intercept) 32.755838 33.381903 33.96577 34.510291 34.891964 34.891964
#> am           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> cyl         -2.028354 -2.132698 -2.23001 -2.320763 -2.384375 -2.384375
#> disp         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#>                1.1855    1.1056    1.0311    0.9616    0.8968    0.8363
#> (Intercept) 34.891964 34.891964 34.891964 34.891964 34.891964 34.891964
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.384375 -2.384375 -2.384375 -2.384375 -2.384375 -2.384375
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                 0.7800    0.7274    0.6784    0.6327    0.5900    0.5503
#> (Intercept) 35.1647678 36.506363 37.015293 37.015185 37.015117 37.015074
#> am           0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.2614058 -1.652602 -1.421300 -1.421173 -1.421094 -1.421044
#> disp         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -0.3275001 -1.945986 -2.560643 -2.560855 -2.560987 -2.561071
#>                0.5132    0.4786    0.4463    0.4163        0.3882      0.3620
#> (Intercept) 37.015047 37.015031 37.015020 37.015014 37.0148772921 37.00269143
#> am           0.000000  0.000000  0.000000  0.000000  0.0000000000  0.00000000
#> carb         0.000000  0.000000  0.000000  0.000000 -0.0004729312 -0.03803158
#> cyl         -1.421012 -1.420993 -1.420980 -1.420972 -1.4207133406 -1.39853238
#> disp         0.000000  0.000000  0.000000  0.000000  0.0000000000  0.00000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.0000000000  0.00000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.0000000000  0.00000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.0000000000  0.00000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.0000000000  0.00000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.0000000000  0.00000000
#> wt          -2.561123 -2.561156 -2.561176 -2.561189 -2.5612330470 -2.56737560
#>                  0.3376     0.3149     0.2937     0.2739     0.2554     0.2382
#> (Intercept) 36.99113026 36.9803359 36.9702684 36.9608793 36.9521231 36.9439570
#> am           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb        -0.07310197 -0.1058213 -0.1363361 -0.1647944 -0.1913346 -0.2160862
#> cyl         -1.37760664 -1.3580741 -1.3398570 -1.3228676 -1.3070233 -1.2922468
#> disp         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> vs           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.57346861 -2.5791693 -2.5844868 -2.5894459 -2.5940708 -2.5983840
#>                 0.2221     0.2072     0.1932     0.1802    0.1680      0.1567
#> (Intercept) 36.9364592 34.4864212 34.4868554 34.4872381 34.487561 33.42342363
#> am           0.0000000  1.4492497  1.4490643  1.4488835  1.448720  1.85051093
#> carb        -0.2390268 -0.5329657 -0.5328752 -0.5328031 -0.532747 -0.28200649
#> cyl         -1.2786347 -1.1829016 -1.1830668 -1.1831764 -1.183247 -0.85173145
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.00000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.00000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.00000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.000000 -0.01572448
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.00000000
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.00000000
#> wt          -2.6022428 -1.9595188 -1.9593893 -1.9593357 -1.959328 -1.80226188
#>                  0.1462      0.1363      0.1271      0.1185      0.1106
#> (Intercept) 33.42273694 33.42220060 33.42178354 28.32895842 28.32610235
#> am           1.85082951  1.85108330  1.85128430  1.28571961  1.28561790
#> carb        -0.28204234 -0.28207181 -0.28209543 -0.50507747 -0.50523192
#> cyl         -0.85155423 -0.85142485 -0.85133098 -0.66116214 -0.66094594
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  1.09236634  1.09289501
#> hp          -0.01572685 -0.01572862 -0.01572994 -0.01998207 -0.01998567
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -1.80229017 -1.80229819 -1.80229366 -1.38687136 -1.38669777
#>                  0.1031      0.0962      0.0897      0.0836      0.0780
#> (Intercept) 30.24644159 30.24566275 30.24496264 30.24434138 30.24379598
#> am           1.49392511  1.49380050  1.49369913  1.49361938  1.49355893
#> carb        -0.55785381 -0.55790796 -0.55795958 -0.55800871 -0.55805524
#> cyl         -0.73917219 -0.73915681 -0.73912683 -0.73908873 -0.73904716
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat        -0.91243610 -0.91279545 -0.91311505 -0.91340009 -0.91365469
#> gear         1.48039657  1.48088925  1.48132338  1.48170500  1.48203970
#> hp          -0.02093173 -0.02093473 -0.02093744 -0.02093985 -0.02094198
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -1.21050349 -1.21026132 -1.21006668 -1.20990841 -1.20977793
#>                  0.0727      0.0678      0.0633      0.0590      0.0550
#> (Intercept) 30.24332147 30.24291176 30.24256024 30.24226023 30.24200528
#> am           1.49351508  1.49348509  1.49346632  1.49345639  1.49345318
#> carb        -0.55809905 -0.55813999 -0.55817798 -0.55821296 -0.55824496
#> cyl         -0.73900537 -0.73896551 -0.73892888 -0.73889617 -0.73886765
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat        -0.91388218 -0.91408534 -0.91426657 -0.91442796 -0.91457142
#> gear         1.48233261  1.48258845  1.48281154  1.48300577  1.48317465
#> hp          -0.02094383 -0.02094544 -0.02094681 -0.02094798 -0.02094897
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -1.20966875 -1.20957601 -1.20949608 -1.20942629 -1.20936468
#>                  0.0513      0.0479      0.0446      0.0416      0.0388
#> (Intercept) 28.97327046 26.29816965 26.30066055 26.30301340 26.30523422
#> am           1.58662900  2.07487688  2.07488797  2.07489780  2.07490673
#> carb        -0.53484374 -0.09590423 -0.09570172 -0.09550912 -0.09532593
#> cyl         -0.60813206 -0.72330064 -0.72355401 -0.72379296 -0.72401826
#> disp         0.00000000  0.01264328  0.01264761  0.01265172  0.01265562
#> drat        -0.81989374 -0.79000656 -0.78997367 -0.78994141 -0.78990990
#> gear         1.48475739  1.19899582  1.19862077  1.19826593  1.19793025
#> hp          -0.02200834 -0.03108593 -0.03108699 -0.03108804 -0.03108908
#> qsec         0.00000000  0.25551442  0.25552699  0.25553877  0.25554991
#> vs           0.38187090  0.35410729  0.35397359  0.35384893  0.35373272
#> wt          -1.20907633 -2.17373347 -2.17409925 -2.17444659 -2.17477656
#>                  0.0362      0.0338      0.0315      0.0294      0.0274
#> (Intercept) 26.30732904 26.30930388 26.31116467 26.31291723 26.31456725
#> am           2.07491507  2.07492306  2.07493090  2.07493876  2.07494676
#> carb        -0.09515169 -0.09498592 -0.09482821 -0.09467814 -0.09453531
#> cyl         -0.72423067 -0.72443090 -0.72461963 -0.72479754 -0.72496525
#> disp         0.01265934  0.01266287  0.01266624  0.01266944  0.01267249
#> drat        -0.78987926 -0.78984957 -0.78982091 -0.78979330 -0.78976680
#> gear         1.19761270  1.19731231  1.19702814  1.19675931  1.19650497
#> hp          -0.03109009 -0.03109109 -0.03109207 -0.03109303 -0.03109396
#> qsec         0.25556051  0.25557069  0.25558051  0.25559005  0.25559936
#> vs           0.35362441  0.35352345  0.35342936  0.35334164  0.35325987
#> wt          -2.17509019 -2.17538843 -2.17567214 -2.17594217 -2.17619926
#>                  0.0255      0.0238      0.0222      0.0207      0.0193
#> (Intercept) 26.31612025 26.31758158 26.31895638 26.32024961 26.32146602
#> am           2.07495498  2.07496350  2.07497236  2.07498158  2.07499117
#> carb        -0.09439937 -0.09426995 -0.09414673 -0.09402939 -0.09391763
#> cyl         -0.72512335 -0.72527242 -0.72541299 -0.72554558 -0.72567066
#> disp         0.01267540  0.01267817  0.01268081  0.01268333  0.01268573
#> drat        -0.78974141 -0.78971715 -0.78969401 -0.78967198 -0.78965105
#> gear         1.19626430  1.19603654  1.19582096  1.19561687  1.19542362
#> hp          -0.03109488 -0.03109577 -0.03109664 -0.03109749 -0.03109831
#> qsec         0.25560847  0.25561743  0.25562625  0.25563496  0.25564357
#> vs           0.35318362  0.35311250  0.35304615  0.35298423  0.35292641
#> wt          -2.17644415 -2.17667750 -2.17689993 -2.17711204 -2.17731436
#>                  0.0180      0.0168      0.0157      0.0146      0.0136
#> (Intercept) 26.32261013 26.32368627 26.32469853 26.32565082 26.32654684
#> am           2.07500115  2.07501148  2.07502216  2.07503316  2.07504446
#> carb        -0.09381116 -0.09370972 -0.09361305 -0.09352091 -0.09343307
#> cyl         -0.72578869 -0.72590010 -0.72600529 -0.72610465 -0.72619852
#> disp         0.01268803  0.01269022  0.01269231  0.01269431  0.01269621
#> drat        -0.78963119 -0.78961237 -0.78959457 -0.78957774 -0.78956186
#> gear         1.19524058  1.19506718  1.19490287  1.19474712  1.19459944
#> hp          -0.03109911 -0.03109989 -0.03110064 -0.03110137 -0.03110208
#> qsec         0.25565209  0.25566051  0.25566885  0.25567710  0.25568527
#> vs           0.35287240  0.35282193  0.35277473  0.35273057  0.35268922
#> wt          -2.17750742 -2.17769169 -2.17786762 -2.17803564 -2.17819615
#>                  0.0127      0.0119      0.0111      0.0103      0.0096
#> (Intercept) 26.32739006 26.32818378 26.32893111 26.32963498 26.33029812
#> am           2.07505600  2.07506778  2.07507973  2.07509184  2.07510406
#> carb        -0.09334932 -0.09326945 -0.09319327 -0.09312059 -0.09305125
#> cyl         -0.72628725 -0.72637115 -0.72645052 -0.72652562 -0.72659673
#> disp         0.01269804  0.01269978  0.01270145  0.01270304  0.01270457
#> drat        -0.78954689 -0.78953278 -0.78951950 -0.78950700 -0.78949526
#> gear         1.19445939  1.19432651  1.19420041  1.19408070  1.19396701
#> hp          -0.03110276 -0.03110343 -0.03110407 -0.03110469 -0.03110528
#> qsec         0.25569334  0.25570131  0.25570918  0.25571694  0.25572459
#> vs           0.35265049  0.35261418  0.35258011  0.35254812  0.35251807
#> wt          -2.17834950 -2.17849607 -2.17863617 -2.17877012 -2.17889822
#>                  0.0090      0.0084      0.0078      0.0073      0.0068
#> (Intercept) 26.33092312 26.33151240 26.33206823 26.33259274 26.33308790
#> am           2.07511636  2.07512870  2.07514105  2.07515338  2.07516565
#> carb        -0.09298507 -0.09292191 -0.09286161 -0.09280405 -0.09274907
#> cyl         -0.72666407 -0.72672789 -0.72678838 -0.72684575 -0.72690018
#> disp         0.01270602  0.01270742  0.01270875  0.01271003  0.01271125
#> drat        -0.78948423 -0.78947388 -0.78946416 -0.78945505 -0.78944650
#> gear         1.19385902  1.19375640  1.19365885  1.19356609  1.19347785
#> hp          -0.03110586 -0.03110642 -0.03110695 -0.03110747 -0.03110796
#> qsec         0.25573211  0.25573951  0.25574677  0.25575390  0.25576088
#> vs           0.35248981  0.35246322  0.35243817  0.35241456  0.35239229
#> wt          -2.17902073 -2.17913792 -2.17925004 -2.17935732 -2.17945998
#>                  0.0063      0.0059      0.0055      0.0051      0.0048
#> (Intercept) 26.33355559 26.33399753 26.33441535 26.33481056 26.33518457
#> am           2.07517784  2.07518993  2.07520188  2.07521369  2.07522532
#> carb        -0.09269657 -0.09264643 -0.09259852 -0.09255274 -0.09250901
#> cyl         -0.72695185 -0.72700092 -0.72704754 -0.72709184 -0.72713397
#> disp         0.01271242  0.01271354  0.01271461  0.01271563  0.01271661
#> drat        -0.78943849 -0.78943099 -0.78942395 -0.78941736 -0.78941119
#> gear         1.19339390  1.19331399  1.19323791  1.19316545  1.19309642
#> hp          -0.03110844 -0.03110890 -0.03110934 -0.03110977 -0.03111018
#> qsec         0.25576771  0.25577440  0.25578093  0.25578730  0.25579352
#> vs           0.35237126  0.35235139  0.35233260  0.35231482  0.35229798
#> wt          -2.17955823 -2.17965228 -2.17974230 -2.17982848 -2.17991098
#>                  0.0045      0.0042
#> (Intercept) 26.33553870 26.33587418
#> am           2.07523676  2.07524800
#> carb        -0.09246721 -0.09242725
#> cyl         -0.72717405 -0.72721218
#> disp         0.01271755  0.01271845
#> drat        -0.78940540 -0.78939998
#> gear         1.19303063  1.19296793
#> hp          -0.03111057 -0.03111095
#> qsec         0.25579957  0.25580546
#> vs           0.35228201  0.35226687
#> wt          -2.17998998 -2.18006561
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1   1  59  85  18   1   1   1   1   1   1   1   2   7
#>  [37]   7   7   7   7   7   7   6  40   1   1   1  40   1   1   1  42   1  38
#>  [55]   1   1   1   1   1   1   1   1   1  68 178   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.162506256 3.881967266 3.620335665 3.376337158 3.148783334 2.936565875
#>   [7] 2.738651164 2.554075241 2.381939118 2.221404394 2.071689172 1.932064256
#>  [13] 1.801849591 1.680410960 1.567156885 1.461535756 1.363033138 1.271169266
#>  [19] 1.185496712 1.105598201 1.031084583 0.961592933 0.896784788 0.836344494
#>  [25] 0.779977675 0.727409790 0.678384805 0.632663939 0.590024507 0.550258830
#>  [31] 0.513173227 0.478587069 0.446331902 0.416250626 0.388196727 0.362033566
#>  [37] 0.337633716 0.314878333 0.293656588 0.273865116 0.255407524 0.238193912
#>  [43] 0.222140439 0.207168917 0.193206426 0.180184959 0.168041096 0.156715689
#>  [49] 0.146153576 0.136303314 0.127116927 0.118549671 0.110559820 0.103108458
#>  [55] 0.096159293 0.089678479 0.083634449 0.077997767 0.072740979 0.067838480
#>  [61] 0.063266394 0.059002451 0.055025883 0.051317323 0.047858707 0.044633190
#>  [67] 0.041625063 0.038819673 0.036203357 0.033763372 0.031487833 0.029365659
#>  [73] 0.027386512 0.025540752 0.023819391 0.022214044 0.020716892 0.019320643
#>  [79] 0.018018496 0.016804110 0.015671569 0.014615358 0.013630331 0.012711693
#>  [85] 0.011854967 0.011055982 0.010310846 0.009615929 0.008967848 0.008363445
#>  [91] 0.007799777 0.007274098 0.006783848 0.006326639 0.005900245 0.005502588
#>  [97] 0.005131732 0.004785871 0.004463319 0.004162506
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 24
#> 
#> $loss
#>   [1] 509.94571 440.09644 381.65713 332.98581 292.66486 259.47112 232.34987
#>   [8] 210.39228 192.81585 178.94741 168.20838 160.10198 154.20215 150.14393
#>  [15] 147.61512 146.34907 146.09009 146.09009 146.09009 146.09009 146.09009
#>  [22] 146.09009 146.09009 146.09009 137.64806 112.85984 110.82504 110.82503
#>  [29] 110.82503 110.82503 110.82503 110.82503 110.82503 110.82503 110.81132
#>  [36] 109.79130 108.96126 108.29348 107.76342 107.34975 107.03411 106.80076
#>  [43] 106.63707 102.70995 102.70995 102.70995 102.70995  99.21581  99.21581
#>  [50]  99.21581  99.21581  96.96066  96.96065  95.98046  95.98045  95.98045
#>  [57]  95.98045  95.98044  95.98044  95.98044  95.98044  95.98044  95.98044
#>  [64]  95.77066  94.04491  94.04489  94.04488  94.04487  94.04486  94.04485
#>  [71]  94.04484  94.04483  94.04482  94.04482  94.04481  94.04481  94.04480
#>  [78]  94.04480  94.04479  94.04479  94.04479  94.04478  94.04478  94.04478
#>  [85]  94.04478  94.04477  94.04477  94.04477  94.04477  94.04477  94.04477
#>  [92]  94.04477  94.04476  94.04476  94.04476  94.04476  94.04476  94.04476
#>  [99]  94.04476  94.04476
#> 
#> $linear.predictors
#>      4.1625   3.8820   3.6203   3.3763   3.1488   2.9366   2.7387   2.5541
#> 1  20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571
#> 2  20.58571 21.06781 21.51742 21.93672 22.32776 22.69245 23.03256 23.34975
#> 3  20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571
#> 4  20.58571 20.10362 19.65401 19.23471 18.84367 18.47898 18.13887 17.82168
#> 5  20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571
#> 6  20.58571 21.06781 21.51742 21.93672 22.32776 22.69245 23.03256 23.34975
#> 7  20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571
#> 8  20.58571 20.10362 19.65401 19.23471 18.84367 18.47898 18.13887 17.82168
#> 9  20.58571 20.10362 19.65401 19.23471 18.84367 18.47898 18.13887 17.82168
#> 10 20.58571 20.10362 19.65401 19.23471 18.84367 18.47898 18.13887 17.82168
#> 11 20.58571 20.10362 19.65401 19.23471 18.84367 18.47898 18.13887 17.82168
#> 12 20.58571 21.06781 21.51742 21.93672 22.32776 22.69245 23.03256 23.34975
#> 13 20.58571 21.06781 21.51742 21.93672 22.32776 22.69245 23.03256 23.34975
#> 14 20.58571 20.10362 19.65401 19.23471 18.84367 18.47898 18.13887 17.82168
#> 15 20.58571 21.06781 21.51742 21.93672 22.32776 22.69245 23.03256 23.34975
#> 16 20.58571 21.06781 21.51742 21.93672 22.32776 22.69245 23.03256 23.34975
#> 17 20.58571 21.06781 21.51742 21.93672 22.32776 22.69245 23.03256 23.34975
#> 18 20.58571 20.10362 19.65401 19.23471 18.84367 18.47898 18.13887 17.82168
#> 19 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571
#> 20 20.58571 20.10362 19.65401 19.23471 18.84367 18.47898 18.13887 17.82168
#> 21 20.58571 21.06781 21.51742 21.93672 22.32776 22.69245 23.03256 23.34975
#>      2.3819   2.2214   2.0717   1.9321   1.8018   1.6804   1.5672   1.4615
#> 1  20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571
#> 2  23.64556 23.92143 24.17871 24.41865 24.64242 24.85111 25.04573 25.22724
#> 3  20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571
#> 4  17.52587 17.25000 16.99272 16.75278 16.52901 16.32032 16.12569 15.94419
#> 5  20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571
#> 6  23.64556 23.92143 24.17871 24.41865 24.64242 24.85111 25.04573 25.22724
#> 7  20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571
#> 8  17.52587 17.25000 16.99272 16.75278 16.52901 16.32032 16.12569 15.94419
#> 9  17.52587 17.25000 16.99272 16.75278 16.52901 16.32032 16.12569 15.94419
#> 10 17.52587 17.25000 16.99272 16.75278 16.52901 16.32032 16.12569 15.94419
#> 11 17.52587 17.25000 16.99272 16.75278 16.52901 16.32032 16.12569 15.94419
#> 12 23.64556 23.92143 24.17871 24.41865 24.64242 24.85111 25.04573 25.22724
#> 13 23.64556 23.92143 24.17871 24.41865 24.64242 24.85111 25.04573 25.22724
#> 14 17.52587 17.25000 16.99272 16.75278 16.52901 16.32032 16.12569 15.94419
#> 15 23.64556 23.92143 24.17871 24.41865 24.64242 24.85111 25.04573 25.22724
#> 16 23.64556 23.92143 24.17871 24.41865 24.64242 24.85111 25.04573 25.22724
#> 17 23.64556 23.92143 24.17871 24.41865 24.64242 24.85111 25.04573 25.22724
#> 18 17.52587 17.25000 16.99272 16.75278 16.52901 16.32032 16.12569 15.94419
#> 19 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571
#> 20 17.52587 17.25000 16.99272 16.75278 16.52901 16.32032 16.12569 15.94419
#> 21 23.64556 23.92143 24.17871 24.41865 24.64242 24.85111 25.04573 25.22724
#>      1.3630   1.2712   1.1855   1.1056   1.0311   0.9616   0.8968   0.8363
#> 1  20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571
#> 2  25.35446 25.35446 25.35446 25.35446 25.35446 25.35446 25.35446 25.35446
#> 3  20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571
#> 4  15.81696 15.81696 15.81696 15.81696 15.81696 15.81696 15.81696 15.81696
#> 5  20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571
#> 6  25.35446 25.35446 25.35446 25.35446 25.35446 25.35446 25.35446 25.35446
#> 7  20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571
#> 8  15.81696 15.81696 15.81696 15.81696 15.81696 15.81696 15.81696 15.81696
#> 9  15.81696 15.81696 15.81696 15.81696 15.81696 15.81696 15.81696 15.81696
#> 10 15.81696 15.81696 15.81696 15.81696 15.81696 15.81696 15.81696 15.81696
#> 11 15.81696 15.81696 15.81696 15.81696 15.81696 15.81696 15.81696 15.81696
#> 12 25.35446 25.35446 25.35446 25.35446 25.35446 25.35446 25.35446 25.35446
#> 13 25.35446 25.35446 25.35446 25.35446 25.35446 25.35446 25.35446 25.35446
#> 14 15.81696 15.81696 15.81696 15.81696 15.81696 15.81696 15.81696 15.81696
#> 15 25.35446 25.35446 25.35446 25.35446 25.35446 25.35446 25.35446 25.35446
#> 16 25.35446 25.35446 25.35446 25.35446 25.35446 25.35446 25.35446 25.35446
#> 17 25.35446 25.35446 25.35446 25.35446 25.35446 25.35446 25.35446 25.35446
#> 18 15.81696 15.81696 15.81696 15.81696 15.81696 15.81696 15.81696 15.81696
#> 19 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571 20.58571
#> 20 15.81696 15.81696 15.81696 15.81696 15.81696 15.81696 15.81696 15.81696
#> 21 25.35446 25.35446 25.35446 25.35446 25.35446 25.35446 25.35446 25.35446
#>      0.7800   0.7274   0.6784   0.6327   0.5900   0.5503   0.5132   0.4786
#> 1  20.65477 20.99604 21.12564 21.12569 21.12572 21.12573 21.12575 21.12575
#> 2  25.35934 25.38127 25.38940 25.38931 25.38925 25.38922 25.38919 25.38918
#> 3  20.54342 20.33440 20.25503 20.25500 20.25498 20.25497 20.25496 20.25496
#> 4  15.94692 16.59135 16.83628 16.83646 16.83657 16.83664 16.83669 16.83671
#> 5  20.46318 19.85764 19.62767 19.62759 19.62754 19.62751 19.62749 19.62748
#> 6  25.08752 23.76610 23.26407 23.26380 23.26363 23.26353 23.26346 23.26342
#> 7  20.46973 19.89656 19.67888 19.67881 19.67876 19.67873 19.67871 19.67870
#> 8  15.74060 15.36538 15.22308 15.22312 15.22315 15.22317 15.22318 15.22319
#> 9  15.85195 16.02702 16.09369 16.09381 16.09389 16.09393 16.09396 16.09398
#> 10 15.83557 15.92972 15.96566 15.96577 15.96584 15.96588 15.96590 15.96592
#> 11 15.32303 12.88425 11.95826 11.95803 11.95789 11.95780 11.95775 11.95771
#> 12 25.39864 25.61478 25.69668 25.69661 25.69657 25.69654 25.69653 25.69652
#> 13 25.31186 25.09910 25.01811 25.01799 25.01791 25.01786 25.01783 25.01781
#> 14 15.94856 16.60108 16.84908 16.84926 16.84938 16.84945 16.84949 16.84952
#> 15 25.48543 26.13047 26.37525 26.37524 26.37523 26.37523 26.37523 26.37522
#> 16 25.41829 25.73154 25.85032 25.85026 25.85023 25.85021 25.85020 25.85019
#> 17 25.62364 26.95168 27.45584 27.45592 27.45597 27.45600 27.45602 27.45603
#> 18 16.03535 17.11677 17.52765 17.52789 17.52804 17.52813 17.52819 17.52823
#> 19 20.68916 21.20037 21.39451 21.39458 21.39462 21.39465 21.39466 21.39467
#> 20 15.90435 16.33837 16.50340 16.50355 16.50364 16.50370 16.50374 16.50376
#> 21 25.20869 24.48611 24.21151 24.21132 24.21120 24.21112 24.21108 24.21105
#>      0.4463   0.4163   0.3882   0.3620   0.3376   0.3149   0.2937   0.2739
#> 1  21.12576 21.12576 21.12516 21.07817 21.03436 20.99349 20.95538 20.91984
#> 2  25.38917 25.38916 25.38949 25.41422 25.43715 25.45855 25.47849 25.49710
#> 3  20.25496 20.25495 20.25576 20.31935 20.37869 20.43404 20.48567 20.53381
#> 4  16.83673 16.83674 16.83758 16.90660 16.97134 17.03176 17.08811 17.14066
#> 5  19.62747 19.62746 19.62826 19.69035 19.74819 19.80214 19.85247 19.89940
#> 6  23.26339 23.26338 23.26319 23.24527 23.22807 23.21201 23.19703 23.18307
#> 7  19.67869 19.67869 19.67806 19.62760 19.58035 19.53626 19.49515 19.45680
#> 8  15.22319 15.22319 15.22353 15.25112 15.27695 15.30106 15.32354 15.34451
#> 9  16.09399 16.09400 16.09435 16.12403 16.15193 16.17798 16.20227 16.22492
#> 10 15.96593 15.96594 15.96629 15.99566 16.02326 16.04902 16.07304 16.09545
#> 11 11.95769 11.95768 11.95749 11.93968 11.92268 11.90680 11.89199 11.87817
#> 12 25.69651 25.69651 25.69684 25.72230 25.74597 25.76805 25.78863 25.80783
#> 13 25.01780 25.01779 25.01811 25.04195 25.06400 25.08457 25.10374 25.12163
#> 14 16.84954 16.84955 16.85039 16.91943 16.98421 17.04465 17.10103 17.15360
#> 15 26.37522 26.37522 26.37557 26.40266 26.42794 26.45153 26.47352 26.49404
#> 16 25.85018 25.85018 25.85004 25.83831 25.82728 25.81697 25.80737 25.79841
#> 17 27.45604 27.45604 27.45593 27.44806 27.44084 27.43411 27.42784 27.42199
#> 18 17.52825 17.52826 17.52817 17.52373 17.51997 17.51649 17.51324 17.51022
#> 19 21.39468 21.39468 21.39314 21.27168 21.15837 21.05266 20.95408 20.86214
#> 20 16.50378 16.50379 16.50179 16.34465 16.19818 16.06154 15.93411 15.81526
#> 21 24.21103 24.21102 24.21085 24.19519 24.18026 24.16631 24.15329 24.14116
#>      0.2554   0.2382   0.2221   0.2072   0.1932   0.1802   0.1680   0.1567
#> 1  20.88669 20.85578 20.82710 21.07278 21.07277 21.07276 21.07275 22.12432
#> 2  25.51445 25.53063 25.54569 26.12502 26.12499 26.12495 26.12491 25.94138
#> 3  20.57871 20.62059 20.65941 20.55619 20.55614 20.55611 20.55609 20.50706
#> 4  17.18966 17.23537 17.27761 17.21653 17.21627 17.21611 17.21601 17.09399
#> 5  19.94316 19.98398 20.02186 20.07611 20.07609 20.07608 20.07606 20.14413
#> 6  23.17004 23.15789 23.14680 22.51640 22.51676 22.51702 22.51720 22.28153
#> 7  19.42104 19.38769 19.35683 18.51640 18.51665 18.51685 18.51701 19.05112
#> 8  15.36406 15.38230 15.39917 15.44907 15.44898 15.44892 15.44888 15.59794
#> 9  16.24605 16.26575 16.28394 16.11531 16.11517 16.11510 16.11505 16.21071
#> 10 16.11634 16.13583 16.15382 16.01733 16.01720 16.01713 16.01709 16.12060
#> 11 11.86529 11.85327 11.84229 12.41772 12.41788 12.41797 12.41799 12.23183
#> 12 25.82574 25.84244 25.85796 26.36016 26.36012 26.36007 26.36003 26.58221
#> 13 25.13831 25.15387 25.16837 24.39164 24.39182 24.39197 24.39208 23.76664
#> 14 17.20263 17.24836 17.29062 17.22633 17.22607 17.22590 17.22580 17.49612
#> 15 26.51317 26.53101 26.54755 26.87943 26.87936 26.87930 26.87925 27.05981
#> 16 25.79005 25.78226 25.77507 25.94476 25.94481 25.94483 25.94484 26.01523
#> 17 27.41653 27.41144 27.40667 27.17338 27.17335 27.17333 27.17334 26.79931
#> 18 17.50739 17.50476 17.50216 18.12892 18.12862 18.12840 18.12825 17.46762
#> 19 20.77640 20.69643 20.62228 20.21260 20.21276 20.21288 20.21298 20.72746
#> 20 15.70443 15.60106 15.50516 15.21325 15.21336 15.21346 15.21353 14.50225
#> 21 24.12984 24.11929 24.10963 24.69067 24.69080 24.69086 24.69087 24.57874
#>      0.1462   0.1363   0.1271   0.1185   0.1106   0.1031   0.0962   0.0897
#> 1  22.12453 22.12469 22.12481 21.81158 21.81152 21.65432 21.65422 21.65415
#> 2  25.94140 25.94142 25.94144 25.75854 25.75848 25.87951 25.87949 25.87946
#> 3  20.50705 20.50704 20.50704 20.47719 20.47722 20.69018 20.69026 20.69034
#> 4  17.09414 17.09424 17.09432 17.03891 17.03902 16.95719 16.95708 16.95701
#> 5  20.14413 20.14412 20.14412 20.23732 20.23741 20.79025 20.79052 20.79075
#> 6  22.28117 22.28090 22.28070 22.77668 22.77670 22.71728 22.71750 22.71765
#> 7  19.05096 19.05084 19.05074 19.48251 19.48260 19.18610 19.18622 19.18631
#> 8  15.59802 15.59808 15.59813 15.56019 15.56024 15.60505 15.60506 15.60507
#> 9  16.21080 16.21086 16.21091 16.03173 16.03172 16.01662 16.01654 16.01649
#> 10 16.12069 16.12075 16.12079 15.96238 15.96239 15.95610 15.95603 15.95599
#> 11 12.23172 12.23165 12.23161 12.28775 12.28769 12.31123 12.31128 12.31130
#> 12 26.58230 26.58237 26.58243 26.46448 26.46450 26.38007 26.38001 26.37996
#> 13 23.76633 23.76609 23.76591 23.09943 23.09895 22.78281 22.78249 22.78219
#> 14 17.49632 17.49647 17.49658 17.54539 17.54560 17.48653 17.48650 17.48649
#> 15 27.05990 27.05998 27.06003 26.83200 26.83197 26.70085 26.70073 26.70063
#> 16 26.01522 26.01522 26.01522 26.63543 26.63572 26.53260 26.53276 26.53290
#> 17 26.79927 26.79923 26.79920 27.06540 27.06549 27.43329 27.43348 27.43365
#> 18 17.46781 17.46796 17.46806 18.09525 18.09565 17.78380 17.78373 17.78368
#> 19 20.72744 20.72743 20.72742 20.74058 20.74048 21.04103 21.04120 21.04135
#> 20 14.50212 14.50202 14.50194 14.10147 14.10106 14.20249 14.20233 14.20218
#> 21 24.57867 24.57863 24.57861 24.29579 24.29559 24.19269 24.19258 24.19246
#>      0.0836   0.0780   0.0727   0.0678   0.0633   0.0590   0.0550   0.0513
#> 1  21.65408 21.65403 21.65398 21.65394 21.65390 21.65386 21.65383 21.61616
#> 2  25.87943 25.87941 25.87939 25.87938 25.87937 25.87936 25.87936 25.90500
#> 3  20.69040 20.69047 20.69052 20.69058 20.69062 20.69067 20.69070 20.79241
#> 4  16.95696 16.95693 16.95692 16.95691 16.95691 16.95691 16.95691 16.89945
#> 5  20.79096 20.79115 20.79132 20.79146 20.79160 20.79171 20.79182 20.86859
#> 6  22.71776 22.71784 22.71789 22.71793 22.71795 22.71796 22.71797 22.67859
#> 7  19.18638 19.18642 19.18645 19.18647 19.18647 19.18647 19.18646 19.42577
#> 8  15.60508 15.60510 15.60512 15.60514 15.60516 15.60517 15.60519 15.55844
#> 9  16.01645 16.01642 16.01641 16.01639 16.01639 16.01638 16.01637 15.96953
#> 10 15.95595 15.95593 15.95592 15.95592 15.95591 15.95591 15.95591 15.90907
#> 11 12.31130 12.31129 12.31128 12.31126 12.31125 12.31123 12.31122 12.25042
#> 12 26.37992 26.37988 26.37984 26.37981 26.37979 26.37977 26.37976 26.45574
#> 13 22.78192 22.78167 22.78145 22.78125 22.78107 22.78091 22.78077 22.69325
#> 14 17.48651 17.48653 17.48656 17.48660 17.48663 17.48666 17.48668 17.45571
#> 15 26.70054 26.70047 26.70041 26.70035 26.70031 26.70027 26.70024 26.77615
#> 16 26.53302 26.53312 26.53320 26.53327 26.53333 26.53338 26.53342 26.25916
#> 17 27.43380 27.43394 27.43406 27.43417 27.43427 27.43436 27.43443 27.45607
#> 18 17.78366 17.78365 17.78366 17.78366 17.78368 17.78369 17.78370 17.87633
#> 19 21.04147 21.04158 21.04167 21.04174 21.04180 21.04185 21.04190 20.95722
#> 20 14.20205 14.20193 14.20182 14.20173 14.20164 14.20157 14.20151 14.24826
#> 21 24.19234 24.19224 24.19213 24.19204 24.19196 24.19190 24.19184 24.24868
#>      0.0479   0.0446   0.0416   0.0388   0.0362   0.0338   0.0315   0.0294
#> 1  22.06743 22.06759 22.06773 22.06788 22.06801 22.06814 22.06827 22.06839
#> 2  25.67905 25.67899 25.67893 25.67887 25.67881 25.67876 25.67871 25.67866
#> 3  21.20150 21.20158 21.20166 21.20174 21.20181 21.20188 21.20194 21.20201
#> 4  17.41118 17.41135 17.41152 17.41167 17.41182 17.41197 17.41210 17.41223
#> 5  20.85924 20.85909 20.85896 20.85882 20.85870 20.85858 20.85847 20.85837
#> 6  23.09745 23.09748 23.09749 23.09751 23.09753 23.09754 23.09756 23.09757
#> 7  19.12173 19.12157 19.12142 19.12129 19.12116 19.12104 19.12093 19.12082
#> 8  14.88613 14.88590 14.88569 14.88549 14.88530 14.88512 14.88495 14.88478
#> 9  15.67630 15.67620 15.67611 15.67603 15.67594 15.67587 15.67579 15.67572
#> 10 15.66982 15.66971 15.66961 15.66951 15.66941 15.66933 15.66924 15.66916
#> 11 12.41915 12.41933 12.41949 12.41965 12.41980 12.41994 12.42007 12.42019
#> 12 26.44681 26.44671 26.44662 26.44653 26.44645 26.44637 26.44629 26.44622
#> 13 22.59485 22.59516 22.59545 22.59572 22.59598 22.59622 22.59645 22.59667
#> 14 17.56272 17.56268 17.56265 17.56261 17.56259 17.56256 17.56253 17.56251
#> 15 26.88100 26.88099 26.88099 26.88098 26.88097 26.88097 26.88096 26.88096
#> 16 26.09075 26.09077 26.09078 26.09079 26.09080 26.09081 26.09082 26.09083
#> 17 27.37780 27.37776 27.37772 27.37768 27.37764 27.37761 27.37758 27.37755
#> 18 17.90951 17.90932 17.90913 17.90896 17.90880 17.90865 17.90850 17.90836
#> 19 20.92544 20.92551 20.92557 20.92562 20.92568 20.92573 20.92578 20.92582
#> 20 14.37989 14.38005 14.38020 14.38034 14.38047 14.38059 14.38071 14.38082
#> 21 24.04226 24.04228 24.04230 24.04231 24.04232 24.04234 24.04235 24.04236
#>      0.0274   0.0255   0.0238   0.0222   0.0207   0.0193   0.0180   0.0168
#> 1  22.06850 22.06861 22.06871 22.06881 22.06891 22.06900 22.06908 22.06916
#> 2  25.67861 25.67856 25.67852 25.67847 25.67843 25.67839 25.67836 25.67832
#> 3  21.20207 21.20213 21.20219 21.20224 21.20230 21.20235 21.20240 21.20244
#> 4  17.41235 17.41247 17.41258 17.41269 17.41279 17.41289 17.41298 17.41307
#> 5  20.85827 20.85818 20.85809 20.85801 20.85793 20.85786 20.85779 20.85772
#> 6  23.09759 23.09760 23.09761 23.09762 23.09764 23.09765 23.09766 23.09767
#> 7  19.12072 19.12063 19.12054 19.12046 19.12039 19.12031 19.12024 19.12018
#> 8  14.88463 14.88448 14.88434 14.88420 14.88407 14.88395 14.88383 14.88372
#> 9  15.67565 15.67559 15.67553 15.67547 15.67542 15.67536 15.67531 15.67526
#> 10 15.66908 15.66901 15.66894 15.66888 15.66881 15.66875 15.66870 15.66864
#> 11 12.42031 12.42042 12.42053 12.42062 12.42072 12.42080 12.42089 12.42097
#> 12 26.44615 26.44609 26.44603 26.44597 26.44592 26.44587 26.44582 26.44578
#> 13 22.59687 22.59706 22.59724 22.59741 22.59757 22.59772 22.59786 22.59800
#> 14 17.56249 17.56247 17.56245 17.56244 17.56242 17.56241 17.56240 17.56239
#> 15 26.88096 26.88095 26.88095 26.88095 26.88095 26.88095 26.88095 26.88094
#> 16 26.09083 26.09084 26.09084 26.09084 26.09084 26.09085 26.09085 26.09084
#> 17 27.37752 27.37750 27.37747 27.37745 27.37742 27.37740 27.37738 27.37736
#> 18 17.90824 17.90811 17.90800 17.90789 17.90779 17.90770 17.90761 17.90752
#> 19 20.92586 20.92591 20.92595 20.92598 20.92602 20.92605 20.92608 20.92611
#> 20 14.38092 14.38102 14.38111 14.38120 14.38128 14.38136 14.38143 14.38150
#> 21 24.04236 24.04237 24.04238 24.04238 24.04239 24.04239 24.04239 24.04239
#>      0.0157   0.0146   0.0136   0.0127   0.0119   0.0111   0.0103   0.0096
#> 1  22.06924 22.06932 22.06939 22.06946 22.06952 22.06959 22.06965 22.06970
#> 2  25.67829 25.67825 25.67822 25.67819 25.67816 25.67814 25.67811 25.67809
#> 3  21.20249 21.20253 21.20258 21.20262 21.20266 21.20269 21.20273 21.20277
#> 4  17.41315 17.41324 17.41331 17.41339 17.41346 17.41353 17.41359 17.41365
#> 5  20.85766 20.85760 20.85755 20.85749 20.85745 20.85740 20.85736 20.85731
#> 6  23.09768 23.09769 23.09770 23.09770 23.09771 23.09772 23.09773 23.09774
#> 7  19.12012 19.12006 19.12000 19.11995 19.11990 19.11985 19.11981 19.11976
#> 8  14.88361 14.88351 14.88341 14.88332 14.88323 14.88314 14.88306 14.88298
#> 9  15.67522 15.67518 15.67513 15.67509 15.67506 15.67502 15.67498 15.67495
#> 10 15.66859 15.66854 15.66850 15.66845 15.66841 15.66837 15.66833 15.66830
#> 11 12.42104 12.42111 12.42118 12.42124 12.42130 12.42135 12.42141 12.42146
#> 12 26.44573 26.44569 26.44565 26.44562 26.44558 26.44555 26.44552 26.44549
#> 13 22.59812 22.59824 22.59835 22.59846 22.59855 22.59865 22.59873 22.59882
#> 14 17.56238 17.56237 17.56236 17.56235 17.56235 17.56234 17.56233 17.56233
#> 15 26.88094 26.88094 26.88094 26.88094 26.88095 26.88095 26.88095 26.88095
#> 16 26.09084 26.09084 26.09084 26.09084 26.09083 26.09083 26.09082 26.09082
#> 17 27.37735 27.37733 27.37731 27.37730 27.37729 27.37727 27.37726 27.37725
#> 18 17.90744 17.90736 17.90729 17.90723 17.90716 17.90710 17.90705 17.90700
#> 19 20.92614 20.92617 20.92619 20.92621 20.92624 20.92626 20.92628 20.92630
#> 20 14.38157 14.38163 14.38170 14.38175 14.38181 14.38186 14.38191 14.38196
#> 21 24.04239 24.04239 24.04239 24.04239 24.04239 24.04239 24.04239 24.04239
#>      0.0090   0.0084   0.0078   0.0073   0.0068   0.0063   0.0059   0.0055
#> 1  22.06976 22.06981 22.06986 22.06991 22.06995 22.07000 22.07004 22.07008
#> 2  25.67806 25.67804 25.67802 25.67800 25.67798 25.67796 25.67794 25.67792
#> 3  21.20280 21.20283 21.20286 21.20289 21.20292 21.20295 21.20298 21.20300
#> 4  17.41371 17.41377 17.41383 17.41388 17.41393 17.41398 17.41402 17.41407
#> 5  20.85728 20.85724 20.85720 20.85717 20.85714 20.85711 20.85708 20.85706
#> 6  23.09775 23.09775 23.09776 23.09777 23.09777 23.09778 23.09779 23.09779
#> 7  19.11972 19.11968 19.11964 19.11961 19.11957 19.11954 19.11951 19.11947
#> 8  14.88291 14.88283 14.88277 14.88270 14.88264 14.88258 14.88252 14.88247
#> 9  15.67492 15.67489 15.67486 15.67483 15.67481 15.67478 15.67476 15.67473
#> 10 15.66826 15.66823 15.66820 15.66817 15.66814 15.66811 15.66808 15.66806
#> 11 12.42150 12.42155 12.42159 12.42163 12.42167 12.42171 12.42174 12.42177
#> 12 26.44546 26.44544 26.44541 26.44539 26.44537 26.44535 26.44533 26.44531
#> 13 22.59890 22.59897 22.59904 22.59910 22.59917 22.59923 22.59928 22.59933
#> 14 17.56232 17.56232 17.56232 17.56231 17.56231 17.56231 17.56230 17.56230
#> 15 26.88095 26.88095 26.88095 26.88095 26.88095 26.88096 26.88096 26.88096
#> 16 26.09082 26.09081 26.09081 26.09080 26.09080 26.09079 26.09079 26.09078
#> 17 27.37724 27.37723 27.37722 27.37721 27.37720 27.37719 27.37718 27.37717
#> 18 17.90695 17.90690 17.90685 17.90681 17.90677 17.90673 17.90670 17.90666
#> 19 20.92631 20.92633 20.92635 20.92636 20.92638 20.92639 20.92640 20.92642
#> 20 14.38200 14.38204 14.38209 14.38213 14.38216 14.38220 14.38223 14.38227
#> 21 24.04239 24.04239 24.04238 24.04238 24.04238 24.04238 24.04237 24.04237
#>      0.0051   0.0048   0.0045   0.0042
#> 1  22.07012 22.07015 22.07019 22.07022
#> 2  25.67790 25.67789 25.67787 25.67786
#> 3  21.20303 21.20305 21.20307 21.20310
#> 4  17.41411 17.41415 17.41419 17.41423
#> 5  20.85703 20.85701 20.85698 20.85696
#> 6  23.09780 23.09781 23.09781 23.09782
#> 7  19.11945 19.11942 19.11939 19.11936
#> 8  14.88241 14.88236 14.88231 14.88227
#> 9  15.67471 15.67469 15.67467 15.67465
#> 10 15.66804 15.66801 15.66799 15.66797
#> 11 12.42181 12.42184 12.42186 12.42189
#> 12 26.44529 26.44527 26.44526 26.44524
#> 13 22.59938 22.59943 22.59948 22.59952
#> 14 17.56230 17.56230 17.56229 17.56229
#> 15 26.88096 26.88096 26.88096 26.88096
#> 16 26.09078 26.09077 26.09077 26.09076
#> 17 27.37717 27.37716 27.37715 27.37715
#> 18 17.90663 17.90660 17.90657 17.90655
#> 19 20.92643 20.92644 20.92645 20.92646
#> 20 14.38230 14.38233 14.38236 14.38238
#> 21 24.04237 24.04237 24.04236 24.04236
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 22.8 21.4 18.7 18.1 22.8 19.2 16.4 17.3 15.2 14.7 32.4 21.5 15.2 27.3
#> [16] 26.0 30.4 15.8 19.7 15.0 21.4
#> 
#> $X
#>               am       carb       cyl       disp       drat       gear
#>  [1,]  1.0488088  0.7236760  0.000000 -0.4812083  0.6319276  0.2399041
#>  [2,]  1.0488088 -0.9649013 -1.145644 -0.9855072  0.5245613  0.2399041
#>  [3,] -0.9534626 -0.9649013  0.000000  0.4692012 -1.1288804 -1.0195922
#>  [4,] -0.9534626 -0.4020422  1.145644  1.4584028 -0.9785675 -1.0195922
#>  [5,] -0.9534626 -0.9649013  0.000000  0.1491653 -1.8160250 -1.0195922
#>  [6,] -0.9534626 -0.4020422 -1.145644 -0.6674109  0.6748741  0.2399041
#>  [7,] -0.9534626  0.7236760  0.000000 -0.4075031  0.6748741  0.2399041
#>  [8,] -0.9534626  0.1608169  1.145644  0.6418265 -1.1503537 -1.0195922
#>  [9,] -0.9534626  0.1608169  1.145644  0.6418265 -1.1503537 -1.0195922
#> [10,] -0.9534626  0.1608169  1.145644  0.6418265 -1.1503537 -1.0195922
#> [11,] -0.9534626  0.7236760  1.145644  2.2342473 -0.8067814 -1.0195922
#> [12,]  1.0488088 -0.9649013 -1.145644 -1.2696602  1.0184464  0.2399041
#> [13,] -0.9534626 -0.9649013 -1.145644 -0.8681607  0.2024622 -1.0195922
#> [14,] -0.9534626 -0.4020422  1.145644  0.9153117 -0.9785675 -1.0195922
#> [15,]  1.0488088 -0.9649013 -1.145644 -1.2667508  1.0184464  0.2399041
#> [16,]  1.0488088 -0.4020422 -1.145644 -0.8662211  1.7700108  1.4994004
#> [17,]  1.0488088 -0.4020422 -1.145644 -1.1106121  0.3527751  1.4994004
#> [18,]  1.0488088  0.7236760  1.145644  1.3711203  1.3190722  1.4994004
#> [19,]  1.0488088  1.8493941  0.000000 -0.6266791  0.0306761  1.4994004
#> [20,]  1.0488088  2.9751123  1.145644  0.8862175 -0.1411100  1.4994004
#> [21,]  1.0488088 -0.4020422 -1.145644 -0.8594324  1.0828662  0.2399041
#>                hp        qsec         vs          wt
#>  [1,] -0.52809334 -0.47865649 -0.9534626 -0.25435598
#>  [2,] -0.78389975  0.36296227  1.0488088 -0.92384986
#>  [3,] -0.52809334  0.80229785  1.0488088  0.15578442
#>  [4,]  0.44998999 -0.47865649 -0.9534626  0.42720086
#>  [5,] -0.60333052  1.21516743  1.0488088  0.45132677
#>  [6,] -0.75380488  2.63374496  1.0488088  0.07737523
#>  [7,] -0.33247668  0.19887308  1.0488088  0.42720086
#>  [8,]  0.52522716 -0.27751490 -0.9534626  1.18716689
#>  [9,]  0.52522716 -0.17165091 -0.9534626  0.77702649
#> [10,]  0.52522716  0.04007708 -0.9534626  0.83734126
#> [11,]  1.27759895 -0.26692850 -0.9534626  2.72519338
#> [12,] -1.19018052  0.81817745  1.0488088 -1.06860530
#> [13,] -0.72371001  1.10401023  1.0488088 -0.74893705
#> [14,]  0.07380409 -0.33044690 -0.9534626  0.42116938
#> [15,] -1.19018052  0.51646506  1.0488088 -1.38827355
#> [16,] -0.81399462 -0.64803888 -0.9534626 -1.14098301
#> [17,] -0.48295103 -0.54217489  1.0488088 -1.89733016
#> [18,]  1.78921177 -1.81254283 -0.9534626  0.10150113
#> [19,]  0.44998999 -1.28322285 -0.9534626 -0.38101698
#> [20,]  2.85757971 -1.75961083 -0.9534626  0.58401925
#> [21,] -0.54314078  0.35766907  1.0488088 -0.36895403
#> attr(,"center")
#>  [1]   0.4761905   2.7142857   6.0000000 209.6190476   3.6057143   3.8095238
#>  [7] 145.0952381  17.9242857   0.4761905   3.0858571
#> attr(,"scale")
#>  [1]   0.4994328   1.7766436   1.7457431 103.1134540   0.4656953   0.7939682
#>  [7]  66.4565055   1.8892164   0.4994328   0.8289844
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 32
#> 
#> $lambda.min
#> [1] 0.4785871
#> 
#> $null.dev
#> [1] 24.28313
#> 
#> $Bias
#> [1] 3.623379
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 8.897417