Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 17.937428 16.866493 15.461320 13.619380 12.116357 10.938505 10.010408
#>   [8]  9.293285  8.753600  8.362518  8.095503  7.931429  7.852295  7.842637
#>  [15]  7.889621  7.906108  7.799510  7.725876  7.751855  7.865336  7.929361
#>  [22]  7.897210  7.781087  7.605343  7.413979  7.281550  7.175171  7.039245
#>  [29]  6.950619  6.933967  6.758841  6.785791  6.838781  6.912290  8.003230
#>  [36]  9.531869  9.242794  8.801266  8.586906  8.531612  8.448186  8.003488
#>  [43]  7.150936  7.044112  7.081901  7.682640  7.407455  7.778815  7.685023
#>  [50]  7.278279  8.024865  8.024908  8.141608  8.544109  8.714252  8.772072
#>  [57]  8.772112  8.848108  8.848078 10.193407 10.976486 11.000421 10.907166
#>  [64] 10.974115 10.974389 10.974631 11.185436 11.334519 11.334972 11.940924
#>  [71] 11.941570 11.942190 11.942785 11.971866 11.972544 12.167414 12.168135
#>  [78] 12.168771 11.864987 11.864736 11.829057 11.891172 11.843834 11.848167
#>  [85] 11.847831 11.847517 11.820365 11.820066 11.819787 11.819525 11.819280
#>  [92] 11.819051 11.818837 11.818637 11.818450 11.818276 11.818113 11.817960
#>  [99] 11.817819 11.817687
#> 
#> $cvse
#>   [1] 5.148358 5.195448 5.021782 4.533302 4.080132 3.677741 3.323329 3.013118
#>   [9] 2.744867 2.517493 2.330622 2.184123 2.077482 2.009229 1.976667 1.956277
#>  [17] 1.945686 1.948862 1.986211 2.045225 2.079324 2.061191 2.046269 2.021992
#>  [25] 1.988912 1.963518 1.945556 1.938335 1.932740 1.933231 1.814872 1.804274
#>  [33] 1.789989 1.773016 1.724307 1.993783 1.876915 1.856660 1.876151 1.850804
#>  [41] 1.831107 1.758200 1.762733 1.785853 1.784297 1.793850 1.767147 1.833321
#>  [49] 1.807572 1.782897 2.095484 2.095456 2.079480 2.295273 2.277321 2.268702
#>  [57] 2.268757 2.257072 2.257073 3.153628 3.466799 3.466529 3.455545 3.449877
#>  [65] 3.450099 3.450315 3.453387 3.433059 3.433236 3.752337 3.752541 3.752740
#>  [73] 3.752932 3.753424 3.753593 3.841790 3.841996 3.842159 3.665058 3.664744
#>  [81] 3.668784 3.678909 3.681347 3.680539 3.680323 3.680120 3.682071 3.681894
#>  [89] 3.681728 3.681572 3.681426 3.681289 3.681161 3.681041 3.680928 3.680823
#>  [97] 3.680724 3.680632 3.680546 3.680466
#> 
#> $fold
#>  [1]  1  4  5  9  4  8  7  3 10  8  2  3  1  2  6  5  6  7  9  1 10
#> 
#> $lambda
#>   [1] 3.597716963 3.355242881 3.129110740 2.918219149 2.721540946 2.538118194
#>   [7] 2.367057523 2.207525768 2.058745920 1.919993335 1.790592210 1.669912288
#>  [13] 1.557365789 1.452404547 1.354517342 1.263227406 1.178090107 1.098690777
#>  [19] 1.024642696 0.955585207 0.891181962 0.831119281 0.775104623 0.722865166
#>  [25] 0.674146473 0.628711257 0.586338222 0.546820989 0.509967084 0.475597009
#>  [31] 0.443543363 0.413650025 0.385771397 0.359771696 0.335524288 0.312911074
#>  [37] 0.291821915 0.272154095 0.253811819 0.236705752 0.220752577 0.205874592
#>  [43] 0.191999334 0.179059221 0.166991229 0.155736579 0.145240455 0.135451734
#>  [49] 0.126322741 0.117809011 0.109869078 0.102464270 0.095558521 0.089118196
#>  [55] 0.083111928 0.077510462 0.072286517 0.067414647 0.062871126 0.058633822
#>  [61] 0.054682099 0.050996708 0.047559701 0.044354336 0.041365002 0.038577140
#>  [67] 0.035977170 0.033552429 0.031291107 0.029182191 0.027215409 0.025381182
#>  [73] 0.023670575 0.022075258 0.020587459 0.019199933 0.017905922 0.016699123
#>  [79] 0.015573658 0.014524045 0.013545173 0.012632274 0.011780901 0.010986908
#>  [85] 0.010246427 0.009555852 0.008911820 0.008311193 0.007751046 0.007228652
#>  [91] 0.006741465 0.006287113 0.005863382 0.005468210 0.005099671 0.004755970
#>  [97] 0.004435434 0.004136500 0.003857714 0.003597717
#> 
#> $fit
#> $beta
#>               3.5977     3.3552     3.1291     2.9182     2.7215     2.5381
#> (Intercept) 18.55238 20.0013031 21.3525727 22.6127713 23.7880367 24.8840932
#> am           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.00000 -0.2204882 -0.4261161 -0.6178855 -0.7967302 -0.9635214
#> disp         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> vs           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#>                2.3671    2.2075    2.0587    1.9200    1.7906    1.6699
#> (Intercept) 25.906279 26.859573 27.748618 28.577745 29.350991 30.072123
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.119071 -1.264138 -1.399427 -1.525599 -1.643267 -1.753004
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.5574   1.4524    1.3545    1.2632   1.1781   1.0987   1.0246
#> (Intercept) 30.744653 31.37186 31.956789 32.502299 32.88467 32.88467 32.88467
#> am           0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.00000
#> carb         0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.00000
#> cyl         -1.855346 -1.95079 -2.039801 -2.122814 -2.18100 -2.18100 -2.18100
#> disp         0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.00000
#> drat         0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.00000
#> gear         0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.00000
#> hp           0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.00000
#> qsec         0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.00000
#> vs           0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.00000
#> wt           0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.00000
#>               0.9556   0.8912   0.8311     0.7751     0.7229     0.6741
#> (Intercept) 32.88467 32.88467 32.88467 33.0806245 33.5179622 33.9257338
#> am           0.00000  0.00000  0.00000  0.0000000  0.0000000  0.0000000
#> carb         0.00000  0.00000  0.00000  0.0000000  0.0000000  0.0000000
#> cyl         -2.18100 -2.18100 -2.18100 -2.1148283 -1.9661346 -1.8275817
#> disp         0.00000  0.00000  0.00000  0.0000000  0.0000000  0.0000000
#> drat         0.00000  0.00000  0.00000  0.0000000  0.0000000  0.0000000
#> gear         0.00000  0.00000  0.00000  0.0000000  0.0000000  0.0000000
#> hp           0.00000  0.00000  0.00000  0.0000000  0.0000000  0.0000000
#> qsec         0.00000  0.00000  0.00000  0.0000000  0.0000000  0.0000000
#> vs           0.00000  0.00000  0.00000  0.0000000  0.0000000  0.0000000
#> wt           0.00000  0.00000  0.00000 -0.1867325 -0.6054502 -0.9956884
#>                0.6287    0.5863    0.5468    0.5100    0.4756      0.4435
#> (Intercept) 34.306156 34.660940 34.985197 34.985050 34.984974 34.99149801
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000 -0.01723207
#> cyl         -1.698192 -1.577522 -1.467056 -1.466969 -1.466924 -1.46209484
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> wt          -1.360006 -1.699771 -2.010648 -2.010775 -2.010840 -2.00807598
#>                  0.4137     0.3858     0.3598     0.3355     0.3129     0.2918
#> (Intercept) 35.00974453 35.0266536 35.0422453 35.0569590 35.0706848 35.0834856
#> am           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb        -0.06217782 -0.1040453 -0.1431724 -0.1795865 -0.2135423 -0.2452094
#> cyl         -1.45023167 -1.4391288 -1.4286275 -1.4189750 -1.4099774 -1.4015864
#> disp         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> vs           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -1.99980760 -1.9921810 -1.9852349 -1.9785956 -1.9723998 -1.9666216
#>                  0.2722      0.2538      0.2367      0.2208      0.2059
#> (Intercept) 33.10427152 32.24118508 31.43483879 30.68460795 26.22963279
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000 -0.59393296
#> cyl         -0.81645592 -0.76626524 -0.71928520 -0.67568429 -0.57498655
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.15832138  0.30799327  0.44779743  0.57790702  1.49435498
#> hp          -0.02291793 -0.02431223 -0.02561642 -0.02682793 -0.02228976
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -1.83563627 -1.77008709 -1.70891931 -1.65191950 -1.20712502
#>                  0.1920      0.1791      0.1670     0.1557      0.1452
#> (Intercept) 26.22879136 26.22816651 26.22770417 26.2273634 26.22711313
#> am           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> carb        -0.59394124 -0.59394444 -0.59394463 -0.5939431 -0.59394084
#> cyl         -0.57481008 -0.57467012 -0.57455989 -0.5744736 -0.57440654
#> disp         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> drat         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> gear         1.49446181  1.49453784  1.49459162  1.4946294  1.49465578
#> hp          -0.02229315 -0.02229581 -0.02229789 -0.0222995 -0.02230075
#> qsec         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> wt          -1.20716869 -1.20721080 -1.20724923 -1.2072830 -1.20731193
#>                 0.1355      0.1263      0.1178     0.1099      0.1025
#> (Intercept) 26.2269301 28.29030789 28.29118144 28.2916503 28.29183798
#> am           0.0000000  0.00000000  0.00000000  0.0000000  0.00000000
#> carb        -0.5939383 -0.64644629 -0.64634026 -0.6462514 -0.64617759
#> cyl         -0.5743546 -0.80247970 -0.80261149 -0.8026741 -0.80268922
#> disp         0.0000000  0.00000000  0.00000000  0.0000000  0.00000000
#> drat         0.0000000  0.00000000  0.00000000  0.0000000  0.00000000
#> gear         1.4946740  1.36299976  1.36293064  1.3628902  1.36286975
#> hp          -0.0223017 -0.02226679 -0.02227122 -0.0222750 -0.02227819
#> qsec         0.0000000  0.00000000  0.00000000  0.0000000  0.00000000
#> vs           0.0000000 -0.86560568 -0.86594686 -0.8661489 -0.86625203
#> wt          -1.2073362 -1.11027428 -1.11005456 -1.1099090 -1.10981810
#>                  0.0956      0.0891      0.0831      0.0775      0.0723
#> (Intercept) 28.29183701 24.90083261 24.89954572 24.89849132 24.89763136
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb        -0.64611686 -0.63210883 -0.63210468 -0.63210188 -0.63210011
#> cyl         -0.80267320 -0.58773705 -0.58754409 -0.58738429 -0.58725245
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         0.00000000  0.92193971  0.92200421  0.92205284  0.92208875
#> gear         1.36286301  1.09308763  1.09315225  1.09320753  1.09325468
#> hp          -0.02228084 -0.02696125 -0.02696254 -0.02696355 -0.02696434
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs          -0.86628662 -0.81051751 -0.81022729 -0.80998450 -0.80978205
#> wt          -1.10976676 -0.99266036 -0.99276163 -0.99284786 -0.99292106
#>                  0.0674      0.0629      0.0586      0.0547      0.0510
#> (Intercept) 24.89693348 24.89637023 24.89591842 24.89555851 24.89527406
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb        -0.63209915 -0.63209878 -0.63209887 -0.63209927 -0.63209990
#> cyl         -0.58714413 -0.58705550 -0.58698333 -0.58692487 -0.58687777
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         0.92211455  0.92213240  0.92214403  0.92215086  0.92215405
#> gear         1.09329478  1.09332879  1.09335755  1.09338179  1.09340217
#> hp          -0.02696495 -0.02696541 -0.02696576 -0.02696601 -0.02696619
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs          -0.80961380 -0.80947449 -0.80935956 -0.80926514 -0.80918789
#> wt          -0.99298303 -0.99303534 -0.99307937 -0.99311631 -0.99314721
#>                   0.0476       0.0444       0.0414       0.0386       0.0360
#> (Intercept) 26.443540814 26.445515964 26.447286211 26.448866706 26.450271731
#> am          -0.525750207 -0.525834386 -0.525914806 -0.525991537 -0.526064655
#> carb        -0.336598046 -0.336600398 -0.336608794 -0.336622524 -0.336640935
#> cyl         -0.779392322 -0.779609353 -0.779806121 -0.779984031 -0.780144411
#> disp         0.009141912  0.009141645  0.009141234  0.009140695  0.009140047
#> drat         0.987249282  0.987038881  0.986847353  0.986673426  0.986515893
#> gear         1.116877761  1.116830063  1.116791137  1.116760172  1.116736417
#> hp          -0.033618044 -0.033614609 -0.033611338 -0.033608226 -0.033605271
#> qsec         0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> vs          -0.606875900 -0.607135417 -0.607382082 -0.607616258 -0.607838319
#> wt          -1.783220796 -1.783219455 -1.783203431 -1.783174429 -1.783134023
#>                   0.0336       0.0313       0.0292       0.0272       0.0254
#> (Intercept) 26.451514743 26.452608398 26.453564594 26.454394499 26.455108582
#> am          -0.526134247 -0.526200403 -0.526263219 -0.526322794 -0.526379231
#> carb        -0.336663424 -0.336689439 -0.336718475 -0.336750069 -0.336783798
#> cyl         -0.780288520 -0.780417544 -0.780532604 -0.780634757 -0.780724999
#> disp         0.009139302  0.009138477  0.009137582  0.009136631  0.009135632
#> drat         0.986373612  0.986245501  0.986130536  0.986027752  0.985936238
#> gear         1.116719171  1.116707783  1.116701650  1.116700214  1.116702956
#> hp          -0.033602466 -0.033599807 -0.033597290 -0.033594910 -0.033592662
#> qsec         0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> vs          -0.608048651 -0.608247648 -0.608435709 -0.608613237 -0.608780634
#> wt          -1.783083658 -1.783024662 -1.782958255 -1.782885550 -1.782807565
#>                   0.0237       0.0221       0.0206       0.0192       0.0179
#> (Intercept) 26.455716651 26.456227881 26.456650845 26.456993542 26.457263429
#> am          -0.526432634 -0.526483110 -0.526530766 -0.526575709 -0.526618047
#> carb        -0.336819279 -0.336856161 -0.336894129 -0.336932897 -0.336972206
#> cyl         -0.780804263 -0.780873431 -0.780933326 -0.780984722 -0.781028343
#> disp         0.009134597  0.009133533  0.009132448  0.009131349  0.009130243
#> drat         0.985855134  0.985783634  0.985720976  0.985666449  0.985619381
#> gear         1.116709400  1.116719105  1.116731664  1.116746707  1.116763890
#> hp          -0.033590541 -0.033588541 -0.033586659 -0.033584889 -0.033583226
#> qsec         0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> vs          -0.608938303 -0.609086643 -0.609226050 -0.609356914 -0.609479619
#> wt          -1.782725227 -1.782639378 -1.782550783 -1.782460132 -1.782368048
#>                   0.0167       0.0156       0.0145       0.0135       0.0126
#> (Intercept) 26.457467450 26.457612059 26.457703249 26.457746577 26.457747189
#> am          -0.526657885 -0.526695330 -0.526730485 -0.526763454 -0.526794336
#> carb        -0.337011827 -0.337051553 -0.337091198 -0.337130601 -0.337169617
#> cyl         -0.781064867 -0.781094924 -0.781119106 -0.781137963 -0.781152006
#> disp         0.009129135  0.009128031  0.009126934  0.009125848  0.009124778
#> drat         0.985579147  0.985545160  0.985516871  0.985493767  0.985475372
#> gear         1.116782900  1.116803452  1.116825284  1.116848158  1.116871860
#> hp          -0.033581666 -0.033580204 -0.033578835 -0.033577555 -0.033576360
#> qsec         0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> vs          -0.609594541 -0.609702049 -0.609802503 -0.609896251 -0.609983636
#> wt          -1.782275090 -1.782181760 -1.782088504 -1.781995719 -1.781903755
#>                   0.0118       0.0110       0.0102      0.0096       0.0089
#> (Intercept) 26.457709844 26.457638933 26.457538504 26.45741228 26.457263696
#> am          -0.526823230 -0.526850232 -0.526875437 -0.52689893 -0.526920811
#> carb        -0.337208119 -0.337245999 -0.337283160 -0.33731952 -0.337355015
#> cyl         -0.781161709 -0.781167516 -0.781169835 -0.78116904 -0.781165495
#> disp         0.009123725  0.009122693  0.009121684  0.00912070  0.009119741
#> drat         0.985461240  0.985450957  0.985444140  0.98544043  0.985439500
#> gear         1.116896194  1.116920984  1.116946072  1.11697131  1.116996586
#> hp          -0.033575245 -0.033574207 -0.033573240 -0.03357234 -0.033571509
#> qsec         0.000000000  0.000000000  0.000000000  0.00000000  0.000000000
#> vs          -0.610064985 -0.610140619 -0.610210846 -0.61027596 -0.610336260
#> wt          -1.781812920 -1.781723482 -1.781635675 -1.78154970 -1.781465720
#>                  0.0083       0.0078       0.0072       0.0067       0.0063
#> (Intercept) 26.45709588 26.456911697 26.456713782 23.902657434 23.898403212
#> am          -0.52694116 -0.526960049 -0.526977571 -0.505667769 -0.505634485
#> carb        -0.33738958 -0.337423177 -0.337455762 -0.291337672 -0.291177351
#> cyl         -0.78115951 -0.781151383 -0.781141392 -0.747440758 -0.747474439
#> disp         0.00911881  0.009117907  0.009117033  0.009878008  0.009881505
#> drat         0.98544104  0.985444775  0.985450438  1.072987793  1.073055688
#> gear         1.11702177  1.117046776  1.117071507  1.134438527  1.134434051
#> hp          -0.03357074 -0.033570022 -0.033569361 -0.033113887 -0.033112571
#> qsec         0.00000000  0.000000000  0.000000000  0.119205118  0.119457760
#> vs          -0.61039201 -0.610443475 -0.610490913 -0.733156957 -0.733386368
#> wt          -1.78138388 -1.781304303 -1.781227075 -1.925743631 -1.926238521
#>                   0.0059       0.0055       0.0051       0.0048       0.0044
#> (Intercept) 23.894284490 23.890297072 23.886436890 23.882700003 23.879082595
#> am          -0.505602498 -0.505571752 -0.505542196 -0.505513778 -0.505486451
#> carb        -0.291022302 -0.290872360 -0.290727366 -0.290587164 -0.290451603
#> cyl         -0.747507273 -0.747539260 -0.747570402 -0.747600704 -0.747630175
#> disp         0.009884887  0.009888157  0.009891318  0.009894374  0.009897329
#> drat         1.073121138  1.073184240  1.073245091  1.073303779  1.073360392
#> gear         1.134429878  1.134425998  1.134422398  1.134419066  1.134415988
#> hp          -0.033111288 -0.033110037 -0.033108817 -0.033107628 -0.033106470
#> qsec         0.119702432  0.119939369  0.120168801  0.120390950  0.120606034
#> vs          -0.733609082 -0.733825259 -0.734035054 -0.734238622 -0.734436117
#> wt          -1.926717394 -1.927180739 -1.927629030 -1.928062731 -1.928482293
#>                   0.0041       0.0039       0.0036
#> (Intercept) 23.875580962 23.872191519 23.868910786
#> am          -0.505460170 -0.505434890 -0.505410570
#> carb        -0.290320536 -0.290193819 -0.290071314
#> cyl         -0.747658821 -0.747686653 -0.747713682
#> disp         0.009900186  0.009902947  0.009905616
#> drat         1.073415010  1.073467711  1.073518569
#> gear         1.134413154  1.134410551  1.134408166
#> hp          -0.033105343 -0.033104245 -0.033103176
#> qsec         0.120814265  0.121015849  0.121210986
#> vs          -0.734627690 -0.734813491 -0.734993669
#> wt          -1.928888154 -1.929280741 -1.929660469
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1  19  22  21  21  21  17   1   1   5   6   6   5   5   5
#>  [37]   5  64  29  29  28  29   1   1   1   1   1   1  68   1   1   1   1  51
#>  [55]   1   1   1   1   1   1   1   1 133   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]  90   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 3.597716963 3.355242881 3.129110740 2.918219149 2.721540946 2.538118194
#>   [7] 2.367057523 2.207525768 2.058745920 1.919993335 1.790592210 1.669912288
#>  [13] 1.557365789 1.452404547 1.354517342 1.263227406 1.178090107 1.098690777
#>  [19] 1.024642696 0.955585207 0.891181962 0.831119281 0.775104623 0.722865166
#>  [25] 0.674146473 0.628711257 0.586338222 0.546820989 0.509967084 0.475597009
#>  [31] 0.443543363 0.413650025 0.385771397 0.359771696 0.335524288 0.312911074
#>  [37] 0.291821915 0.272154095 0.253811819 0.236705752 0.220752577 0.205874592
#>  [43] 0.191999334 0.179059221 0.166991229 0.155736579 0.145240455 0.135451734
#>  [49] 0.126322741 0.117809011 0.109869078 0.102464270 0.095558521 0.089118196
#>  [55] 0.083111928 0.077510462 0.072286517 0.067414647 0.062871126 0.058633822
#>  [61] 0.054682099 0.050996708 0.047559701 0.044354336 0.041365002 0.038577140
#>  [67] 0.035977170 0.033552429 0.031291107 0.029182191 0.027215409 0.025381182
#>  [73] 0.023670575 0.022075258 0.020587459 0.019199933 0.017905922 0.016699123
#>  [79] 0.015573658 0.014524045 0.013545173 0.012632274 0.011780901 0.010986908
#>  [85] 0.010246427 0.009555852 0.008911820 0.008311193 0.007751046 0.007228652
#>  [91] 0.006741465 0.006287113 0.005863382 0.005468210 0.005099671 0.004755970
#>  [97] 0.004435434 0.004136500 0.003857714 0.003597717
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 22
#> 
#> $loss
#>   [1] 367.37238 315.19214 271.53560 235.17615 205.05477 180.25770 159.99702
#>   [8] 143.59381 130.46351 120.10322 112.08073 106.02492 101.61751  98.58585
#>  [15]  96.69673  95.75093  95.55747  95.55747  95.55747  95.55747  95.55747
#>  [22]  95.55747  90.67760  81.46275  75.02580  70.88961  68.66227  68.00261
#>  [29]  68.00261  68.00261  67.57012  66.52969  65.67450  64.97468  64.40965
#>  [36]  63.95770  63.60139  56.53681  55.58778  54.92907  54.51366  49.71384
#>  [43]  49.71384  49.71383  49.71383  49.71383  49.71383  49.71383  48.60550
#>  [50]  48.60550  48.60550  48.60550  48.60550  47.70535  47.70534  47.70534
#>  [57]  47.70534  47.70534  47.70534  47.70534  47.70534  47.70534  46.40485
#>  [64]  46.40485  46.40485  46.40484  46.40484  46.40484  46.40484  46.40483
#>  [71]  46.40483  46.40483  46.40483  46.40483  46.40483  46.40483  46.40483
#>  [78]  46.40482  46.40482  46.40482  46.40482  46.40482  46.40482  46.40482
#>  [85]  46.40482  46.40482  46.40482  46.40482  46.40482  46.40482  46.37602
#>  [92]  46.37601  46.37600  46.37599  46.37598  46.37597  46.37596  46.37596
#>  [99]  46.37595  46.37594
#> 
#> $linear.predictors
#>      3.5977   3.3552   3.1291   2.9182   2.7215   2.5381   2.3671   2.2075
#> 1  18.55238 18.67837 18.79588 18.90546 19.00766 19.10296 19.19185 19.27475
#> 2  18.55238 18.67837 18.79588 18.90546 19.00766 19.10296 19.19185 19.27475
#> 3  18.55238 18.23740 17.94364 17.66969 17.41419 17.17592 16.95371 16.74647
#> 4  18.55238 18.67837 18.79588 18.90546 19.00766 19.10296 19.19185 19.27475
#> 5  18.55238 18.23740 17.94364 17.66969 17.41419 17.17592 16.95371 16.74647
#> 6  18.55238 19.11935 19.64811 20.14123 20.60112 21.03001 21.42999 21.80302
#> 7  18.55238 18.67837 18.79588 18.90546 19.00766 19.10296 19.19185 19.27475
#> 8  18.55238 18.23740 17.94364 17.66969 17.41419 17.17592 16.95371 16.74647
#> 9  18.55238 18.23740 17.94364 17.66969 17.41419 17.17592 16.95371 16.74647
#> 10 18.55238 18.23740 17.94364 17.66969 17.41419 17.17592 16.95371 16.74647
#> 11 18.55238 18.23740 17.94364 17.66969 17.41419 17.17592 16.95371 16.74647
#> 12 18.55238 19.11935 19.64811 20.14123 20.60112 21.03001 21.42999 21.80302
#> 13 18.55238 18.23740 17.94364 17.66969 17.41419 17.17592 16.95371 16.74647
#> 14 18.55238 18.23740 17.94364 17.66969 17.41419 17.17592 16.95371 16.74647
#> 15 18.55238 18.23740 17.94364 17.66969 17.41419 17.17592 16.95371 16.74647
#> 16 18.55238 18.23740 17.94364 17.66969 17.41419 17.17592 16.95371 16.74647
#> 17 18.55238 19.11935 19.64811 20.14123 20.60112 21.03001 21.42999 21.80302
#> 18 18.55238 19.11935 19.64811 20.14123 20.60112 21.03001 21.42999 21.80302
#> 19 18.55238 18.23740 17.94364 17.66969 17.41419 17.17592 16.95371 16.74647
#> 20 18.55238 18.67837 18.79588 18.90546 19.00766 19.10296 19.19185 19.27475
#> 21 18.55238 19.11935 19.64811 20.14123 20.60112 21.03001 21.42999 21.80302
#>      2.0587   1.9200   1.7906   1.6699   1.5574   1.4524   1.3545   1.2632
#> 1  19.35205 19.42415 19.49139 19.55410 19.61258 19.66712 19.71798 19.76542
#> 2  19.35205 19.42415 19.49139 19.55410 19.61258 19.66712 19.71798 19.76542
#> 3  16.55320 16.37295 16.20486 16.04809 15.90189 15.76554 15.63838 15.51979
#> 4  19.35205 19.42415 19.49139 19.55410 19.61258 19.66712 19.71798 19.76542
#> 5  16.55320 16.37295 16.20486 16.04809 15.90189 15.76554 15.63838 15.51979
#> 6  22.15091 22.47535 22.77792 23.06011 23.32327 23.56870 23.79758 24.01104
#> 7  19.35205 19.42415 19.49139 19.55410 19.61258 19.66712 19.71798 19.76542
#> 8  16.55320 16.37295 16.20486 16.04809 15.90189 15.76554 15.63838 15.51979
#> 9  16.55320 16.37295 16.20486 16.04809 15.90189 15.76554 15.63838 15.51979
#> 10 16.55320 16.37295 16.20486 16.04809 15.90189 15.76554 15.63838 15.51979
#> 11 16.55320 16.37295 16.20486 16.04809 15.90189 15.76554 15.63838 15.51979
#> 12 22.15091 22.47535 22.77792 23.06011 23.32327 23.56870 23.79758 24.01104
#> 13 16.55320 16.37295 16.20486 16.04809 15.90189 15.76554 15.63838 15.51979
#> 14 16.55320 16.37295 16.20486 16.04809 15.90189 15.76554 15.63838 15.51979
#> 15 16.55320 16.37295 16.20486 16.04809 15.90189 15.76554 15.63838 15.51979
#> 16 16.55320 16.37295 16.20486 16.04809 15.90189 15.76554 15.63838 15.51979
#> 17 22.15091 22.47535 22.77792 23.06011 23.32327 23.56870 23.79758 24.01104
#> 18 22.15091 22.47535 22.77792 23.06011 23.32327 23.56870 23.79758 24.01104
#> 19 16.55320 16.37295 16.20486 16.04809 15.90189 15.76554 15.63838 15.51979
#> 20 19.35205 19.42415 19.49139 19.55410 19.61258 19.66712 19.71798 19.76542
#> 21 22.15091 22.47535 22.77792 23.06011 23.32327 23.56870 23.79758 24.01104
#>      1.1781   1.0987   1.0246   0.9556   0.8912   0.8311   0.7751   0.7229
#> 1  19.79867 19.79867 19.79867 19.79867 19.79867 19.79867 19.90242 20.13488
#> 2  19.79867 19.79867 19.79867 19.79867 19.79867 19.79867 19.79131 19.77463
#> 3  15.43667 15.43667 15.43667 15.43667 15.43667 15.43667 15.51964 15.70614
#> 4  19.79867 19.79867 19.79867 19.79867 19.79867 19.79867 19.74556 19.62630
#> 5  15.43667 15.43667 15.43667 15.43667 15.43667 15.43667 15.49536 15.62743
#> 6  24.16067 24.16067 24.16067 24.16067 24.16067 24.16067 24.02563 23.72204
#> 7  19.79867 19.79867 19.79867 19.79867 19.79867 19.79867 19.74929 19.63841
#> 8  15.43667 15.43667 15.43667 15.43667 15.43667 15.43667 15.46549 15.53056
#> 9  15.43667 15.43667 15.43667 15.43667 15.43667 15.43667 15.45615 15.50028
#> 10 15.43667 15.43667 15.43667 15.43667 15.43667 15.43667 15.18165 14.61027
#> 11 15.43667 15.43667 15.43667 15.43667 15.43667 15.43667 15.16391 14.55275
#> 12 24.16067 24.16067 24.16067 24.16067 24.16067 24.16067 24.16102 24.16099
#> 13 15.43667 15.43667 15.43667 15.43667 15.43667 15.43667 15.50470 15.65770
#> 14 15.43667 15.43667 15.43667 15.43667 15.43667 15.43667 15.52057 15.70916
#> 15 15.43667 15.43667 15.43667 15.43667 15.43667 15.43667 15.44494 15.46396
#> 16 15.43667 15.43667 15.43667 15.43667 15.43667 15.43667 15.44401 15.46093
#> 17 24.16067 24.16067 24.16067 24.16067 24.16067 24.16067 24.25998 24.48188
#> 18 24.16067 24.16067 24.16067 24.16067 24.16067 24.16067 24.22170 24.35776
#> 19 15.43667 15.43667 15.43667 15.43667 15.43667 15.43667 15.57006 15.86961
#> 20 19.79867 19.79867 19.79867 19.79867 19.79867 19.79867 19.87441 20.04406
#> 21 24.16067 24.16067 24.16067 24.16067 24.16067 24.16067 24.10219 23.97027
#>      0.6741   0.6287   0.5863   0.5468   0.5100   0.4756   0.4435   0.4137
#> 1  20.35154 20.55379 20.74241 20.91496 20.91501 20.91503 20.88884 20.82015
#> 2  19.75911 19.74459 19.73105 19.71863 19.71860 19.71858 19.74573 19.81680
#> 3  15.87991 16.04220 16.19355 16.33212 16.33223 16.33230 16.35249 16.40420
#> 4  19.51516 19.41139 19.31460 19.22602 19.22596 19.22593 19.25375 19.32684
#> 5  15.75047 15.86540 15.97258 16.07073 16.07083 16.07089 16.05698 16.01987
#> 6  23.43916 23.17497 22.92858 22.70300 22.70280 22.70270 22.70289 22.70508
#> 7  19.53508 19.43859 19.34860 19.26623 19.26617 19.26614 19.24222 19.18031
#> 8  15.59116 15.64780 15.70062 15.74903 15.74911 15.74915 15.75292 15.76208
#> 9  15.54138 15.57980 15.61563 15.64850 15.64857 15.64861 15.65252 15.66208
#> 10 14.07772 13.58059 13.11697 12.69284 12.69273 12.69267 12.68341 12.66019
#> 11 13.98313 13.45139 12.95549 12.50183 12.50171 12.50164 12.49264 12.47021
#> 12 24.16104 24.16097 24.16092 24.16072 24.16062 24.16056 24.17598 24.21711
#> 13 15.80026 15.93340 16.05757 16.17127 16.17137 16.17143 16.19185 16.24421
#> 14 15.88489 16.04900 16.20205 16.34217 16.34229 16.34235 16.36253 16.41420
#> 15 15.48164 15.49820 15.51365 15.52786 15.52792 15.52796 15.51480 15.47992
#> 16 15.47666 15.49140 15.50515 15.51781 15.51787 15.51791 15.53922 15.59428
#> 17 24.68875 24.88178 25.06180 25.22637 25.22633 25.22630 25.24026 25.27701
#> 18 24.48463 24.60298 24.71334 24.81419 24.81412 24.81408 24.81137 24.80487
#> 19 16.14875 16.40940 16.65249 16.87499 16.87514 16.87522 16.86021 16.81979
#> 20 20.20219 20.34979 20.48744 20.61336 20.61339 20.61341 20.55317 20.39582
#> 21 23.84739 23.73257 23.62549 23.52737 23.52722 23.52714 23.52620 23.52500
#>      0.3858   0.3598   0.3355   0.3129   0.2918   0.2722   0.2538   0.2367
#> 1  20.75619 20.69648 20.64084 20.58896 20.54058 21.50848 21.56359 21.61514
#> 2  19.88297 19.94478 20.00234 20.05601 20.10607 20.25796 20.20240 20.15054
#> 3  16.45243 16.49767 16.53962 16.57873 16.61520 16.72236 16.69130 16.66239
#> 4  19.39489 19.45840 19.51758 19.57277 19.62425 19.92282 19.89029 19.85993
#> 5  15.98536 15.95325 15.92323 15.89523 15.86912 14.87947 14.75934 14.64708
#> 6  22.70699 22.70849 22.71017 22.71174 22.71320 23.19514 23.25416 23.30922
#> 7  19.12260 19.06858 19.01839 18.97160 18.92795 19.70533 19.79606 19.88081
#> 8  15.77065 15.77878 15.78624 15.79319 15.79967 16.07544 16.05642 16.03872
#> 9  15.67104 15.67952 15.68731 15.69457 15.70134 15.98366 15.96791 15.95328
#> 10 12.63849 12.61805 12.59919 12.58160 12.56519 12.71232 12.75808 12.80076
#> 11 12.44923 12.42946 12.41122 12.39422 12.37836 11.96499 11.98211 11.99800
#> 12 24.25537 24.29096 24.32423 24.35527 24.38421 23.56553 23.37855 23.20381
#> 13 16.29306 16.33885 16.38133 16.42093 16.45787 17.14846 17.15750 17.16609
#> 14 16.46239 16.50760 16.54951 16.58859 16.62503 17.30449 17.30796 17.31135
#> 15 15.44747 15.41723 15.38901 15.36268 15.33813 14.38385 14.28141 14.18568
#> 16 15.64560 15.69365 15.73829 15.77990 15.81872 15.97893 15.97442 15.97028
#> 17 25.31122 25.34313 25.37289 25.40064 25.42652 25.40719 25.37837 25.35144
#> 18 24.79878 24.79299 24.78769 24.78275 24.77815 24.61626 24.71569 24.80850
#> 19 16.78223 16.74734 16.71466 16.68419 16.65577 15.49493 15.62142 15.73953
#> 20 20.24927 20.11235 19.98488 19.86602 19.75517 19.90179 20.02578 20.14153
#> 21 23.52378 23.52244 23.52139 23.52042 23.51951 22.87061 22.83722 22.80590
#>      0.2208   0.2059   0.1920   0.1791   0.1670   0.1557   0.1452   0.1355
#> 1  21.66303 20.76686 20.76698 20.76709 20.76717 20.76724 20.76729 20.76734
#> 2  20.10223 20.33606 20.33608 20.33609 20.33610 20.33611 20.33611 20.33611
#> 3  16.63536 16.87172 16.87185 16.87196 16.87204 16.87211 16.87216 16.87220
#> 4  19.83165 20.15177 20.15179 20.15181 20.15181 20.15182 20.15182 20.15182
#> 5  14.54266 13.96665 13.96652 13.96642 13.96636 13.96631 13.96628 13.96626
#> 6  23.36054 23.48655 23.48647 23.48641 23.48635 23.48630 23.48626 23.48622
#> 7  19.95969 19.48725 19.48730 19.48733 19.48735 19.48737 19.48739 19.48740
#> 8  16.02217 15.81627 15.81637 15.81644 15.81651 15.81656 15.81660 15.81663
#> 9  15.93957 15.75592 15.75601 15.75608 15.75614 15.75619 15.75623 15.75626
#> 10 12.84055 12.83027 12.83020 12.83014 12.83010 12.83006 12.83002 12.83000
#> 11 12.01292 12.15835 12.15819 12.15806 12.15796 12.15788 12.15781 12.15776
#> 12 23.04130 22.68115 22.68089 22.68069 22.68053 22.68041 22.68032 22.68025
#> 13 17.17391 17.33240 17.33261 17.33278 17.33291 17.33301 17.33310 17.33316
#> 14 17.31432 17.43500 17.43522 17.43539 17.43553 17.43563 17.43572 17.43578
#> 15 14.09664 13.64072 13.64058 13.64048 13.64040 13.64034 13.64030 13.64027
#> 16 15.96634 16.38284 16.38295 16.38304 16.38311 16.38316 16.38320 16.38323
#> 17 25.32639 25.50626 25.50624 25.50622 25.50620 25.50618 25.50617 25.50616
#> 18 24.89496 25.60198 25.60196 25.60194 25.60191 25.60189 25.60187 25.60185
#> 19 15.84951 17.01470 17.01474 17.01476 17.01478 17.01479 17.01480 17.01481
#> 20 20.24933 19.44345 19.44343 19.44343 19.44342 19.44342 19.44342 19.44342
#> 21 22.77692 22.93385 22.93363 22.93346 22.93332 22.93321 22.93312 22.93305
#>      0.1263   0.1178   0.1099   0.1025   0.0956   0.0891   0.0831   0.0775
#> 1  20.98338 20.98370 20.98395 20.98415 20.98430 21.24738 21.24737 21.24736
#> 2  20.03350 20.03336 20.03327 20.03322 20.03319 19.89348 19.89357 19.89365
#> 3  16.95055 16.95035 16.95021 16.95012 16.95006 16.98512 16.98521 16.98528
#> 4  19.87282 19.87275 19.87272 19.87270 19.87270 19.49006 19.49012 19.49016
#> 5  13.95464 13.95438 13.95417 13.95402 13.95390 13.75989 13.75988 13.75989
#> 6  23.45157 23.45194 23.45220 23.45239 23.45251 23.41127 23.41115 23.41105
#> 7  18.91788 18.91798 18.91810 18.91823 18.91835 19.29083 19.29101 19.29116
#> 8  15.87079 15.87074 15.87071 15.87070 15.87071 15.85658 15.85663 15.85667
#> 9  15.81527 15.81523 15.81522 15.81521 15.81522 15.80695 15.80699 15.80703
#> 10 12.98005 12.98033 12.98052 12.98065 12.98072 12.91252 12.91238 12.91227
#> 11 12.31791 12.31810 12.31821 12.31826 12.31828 12.42077 12.42061 12.42047
#> 12 22.76063 22.76065 22.76062 22.76057 22.76051 22.73555 22.73539 22.73525
#> 13 17.41839 17.41833 17.41830 17.41829 17.41830 17.22018 17.22027 17.22034
#> 14 17.51277 17.51268 17.51264 17.51262 17.51263 17.66412 17.66424 17.66434
#> 15 13.65487 13.65466 13.65450 13.65436 13.65426 13.97128 13.97128 13.97128
#> 16 16.50088 16.50078 16.50070 16.50064 16.50061 16.51856 16.51860 16.51864
#> 17 25.40235 25.40231 25.40229 25.40227 25.40226 25.54088 25.54090 25.54092
#> 18 26.20023 26.20051 26.20067 26.20075 26.20077 26.25753 26.25730 26.25710
#> 19 16.70168 16.70111 16.70069 16.70040 16.70020 16.76202 16.76223 16.76241
#> 20 19.43960 19.43981 19.43997 19.44011 19.44021 18.91673 18.91668 18.91663
#> 21 22.86025 22.86032 22.86034 22.86034 22.86031 22.93830 22.93819 22.93809
#>      0.0723   0.0674   0.0629   0.0586   0.0547   0.0510   0.0476   0.0444
#> 1  21.24735 21.24734 21.24733 21.24731 21.24730 21.24729 21.30551 21.30542
#> 2  19.89371 19.89377 19.89381 19.89385 19.89388 19.89391 20.14265 20.14258
#> 3  16.98535 16.98540 16.98544 16.98548 16.98551 16.98553 17.26933 17.26926
#> 4  19.49020 19.49024 19.49027 19.49029 19.49031 19.49033 19.25625 19.25624
#> 5  13.75989 13.75989 13.75990 13.75991 13.75991 13.75992 14.07028 14.07044
#> 6  23.41096 23.41088 23.41082 23.41077 23.41073 23.41069 23.72469 23.72474
#> 7  19.29128 19.29138 19.29145 19.29152 19.29157 19.29160 19.41433 19.41410
#> 8  15.85671 15.85674 15.85677 15.85679 15.85680 15.85682 15.39877 15.39877
#> 9  15.80706 15.80709 15.80711 15.80713 15.80715 15.80716 15.30961 15.30961
#> 10 12.91217 12.91209 12.91202 12.91196 12.91192 12.91188 13.16666 13.16671
#> 11 12.42036 12.42027 12.42019 12.42012 12.42007 12.42002 12.16043 12.16052
#> 12 22.73514 22.73505 22.73497 22.73491 22.73486 22.73482 22.82731 22.82754
#> 13 17.22041 17.22046 17.22050 17.22054 17.22057 17.22060 17.19813 17.19808
#> 14 17.66442 17.66449 17.66454 17.66459 17.66463 17.66465 17.60675 17.60661
#> 15 13.97129 13.97129 13.97129 13.97129 13.97129 13.97129 14.01076 14.01082
#> 16 16.51867 16.51869 16.51871 16.51873 16.51875 16.51876 16.84369 16.84363
#> 17 25.54093 25.54094 25.54095 25.54096 25.54096 25.54096 25.40512 25.40504
#> 18 26.25694 26.25681 26.25669 26.25660 26.25652 26.25646 26.30937 26.30950
#> 19 16.76255 16.76268 16.76278 16.76287 16.76294 16.76300 16.76768 16.76751
#> 20 18.91660 18.91657 18.91655 18.91654 18.91652 18.91652 18.88298 18.88312
#> 21 22.93801 22.93794 22.93789 22.93784 22.93780 22.93777 22.52971 22.52975
#>      0.0414   0.0386   0.0360   0.0336   0.0313   0.0292   0.0272   0.0254
#> 1  21.30533 21.30524 21.30515 21.30507 21.30499 21.30491 21.30483 21.30476
#> 2  20.14251 20.14245 20.14238 20.14231 20.14225 20.14219 20.14212 20.14206
#> 3  17.26920 17.26914 17.26908 17.26902 17.26897 17.26891 17.26886 17.26880
#> 4  19.25623 19.25623 19.25623 19.25623 19.25624 19.25624 19.25625 19.25626
#> 5  14.07059 14.07071 14.07081 14.07091 14.07098 14.07105 14.07110 14.07114
#> 6  23.72479 23.72484 23.72488 23.72493 23.72497 23.72500 23.72504 23.72507
#> 7  19.41389 19.41371 19.41353 19.41338 19.41324 19.41311 19.41300 19.41290
#> 8  15.39877 15.39878 15.39879 15.39881 15.39884 15.39887 15.39891 15.39895
#> 9  15.30961 15.30962 15.30964 15.30966 15.30969 15.30972 15.30976 15.30981
#> 10 13.16676 13.16679 13.16682 13.16684 13.16685 13.16686 13.16686 13.16686
#> 11 12.16061 12.16069 12.16077 12.16084 12.16091 12.16098 12.16105 12.16112
#> 12 22.82775 22.82794 22.82812 22.82828 22.82843 22.82856 22.82869 22.82880
#> 13 17.19803 17.19798 17.19794 17.19790 17.19787 17.19784 17.19781 17.19778
#> 14 17.60649 17.60638 17.60629 17.60620 17.60612 17.60605 17.60599 17.60594
#> 15 14.01087 14.01091 14.01096 14.01099 14.01103 14.01106 14.01109 14.01111
#> 16 16.84358 16.84352 16.84346 16.84340 16.84334 16.84329 16.84323 16.84318
#> 17 25.40497 25.40490 25.40483 25.40477 25.40471 25.40466 25.40461 25.40456
#> 18 26.30962 26.30974 26.30986 26.30997 26.31007 26.31017 26.31026 26.31035
#> 19 16.76736 16.76722 16.76710 16.76698 16.76687 16.76677 16.76669 16.76660
#> 20 18.88325 18.88336 18.88347 18.88356 18.88365 18.88372 18.88379 18.88385
#> 21 22.52980 22.52985 22.52989 22.52994 22.52998 22.53002 22.53006 22.53010
#>      0.0237   0.0221   0.0206   0.0192   0.0179   0.0167   0.0156   0.0145
#> 1  21.30469 21.30462 21.30456 21.30450 21.30444 21.30438 21.30433 21.30428
#> 2  20.14200 20.14194 20.14188 20.14183 20.14178 20.14172 20.14167 20.14163
#> 3  17.26875 17.26870 17.26865 17.26861 17.26856 17.26852 17.26848 17.26844
#> 4  19.25627 19.25628 19.25629 19.25630 19.25632 19.25633 19.25634 19.25635
#> 5  14.07117 14.07120 14.07121 14.07122 14.07122 14.07122 14.07121 14.07120
#> 6  23.72510 23.72513 23.72516 23.72519 23.72521 23.72523 23.72525 23.72527
#> 7  19.41281 19.41273 19.41267 19.41261 19.41255 19.41251 19.41247 19.41244
#> 8  15.39899 15.39903 15.39908 15.39912 15.39917 15.39922 15.39927 15.39932
#> 9  15.30985 15.30990 15.30995 15.31000 15.31005 15.31011 15.31016 15.31021
#> 10 13.16685 13.16684 13.16682 13.16681 13.16679 13.16676 13.16674 13.16671
#> 11 12.16118 12.16124 12.16129 12.16135 12.16140 12.16145 12.16149 12.16154
#> 12 22.82890 22.82899 22.82907 22.82915 22.82921 22.82927 22.82932 22.82937
#> 13 17.19776 17.19774 17.19772 17.19771 17.19769 17.19768 17.19767 17.19766
#> 14 17.60589 17.60585 17.60582 17.60579 17.60576 17.60574 17.60572 17.60571
#> 15 14.01113 14.01115 14.01117 14.01119 14.01120 14.01121 14.01122 14.01123
#> 16 16.84312 16.84307 16.84302 16.84297 16.84292 16.84287 16.84283 16.84278
#> 17 25.40452 25.40448 25.40444 25.40441 25.40438 25.40435 25.40432 25.40430
#> 18 26.31043 26.31051 26.31059 26.31066 26.31073 26.31079 26.31085 26.31091
#> 19 16.76653 16.76646 16.76640 16.76635 16.76630 16.76626 16.76622 16.76618
#> 20 18.88391 18.88396 18.88400 18.88403 18.88406 18.88409 18.88411 18.88413
#> 21 22.53013 22.53017 22.53020 22.53023 22.53026 22.53029 22.53032 22.53035
#>      0.0135   0.0126   0.0118   0.0110   0.0102   0.0096   0.0089   0.0083
#> 1  21.30423 21.30418 21.30414 21.30409 21.30406 21.30402 21.30398 21.30395
#> 2  20.14158 20.14153 20.14149 20.14145 20.14141 20.14137 20.14134 20.14130
#> 3  17.26840 17.26836 17.26833 17.26829 17.26826 17.26823 17.26820 17.26818
#> 4  19.25637 19.25638 19.25639 19.25641 19.25642 19.25643 19.25644 19.25646
#> 5  14.07118 14.07116 14.07114 14.07111 14.07108 14.07106 14.07102 14.07099
#> 6  23.72529 23.72531 23.72532 23.72533 23.72535 23.72536 23.72537 23.72538
#> 7  19.41242 19.41240 19.41238 19.41237 19.41236 19.41236 19.41235 19.41236
#> 8  15.39937 15.39942 15.39946 15.39951 15.39956 15.39960 15.39965 15.39969
#> 9  15.31027 15.31032 15.31037 15.31043 15.31048 15.31053 15.31058 15.31062
#> 10 13.16669 13.16666 13.16663 13.16660 13.16657 13.16654 13.16651 13.16648
#> 11 12.16158 12.16162 12.16165 12.16169 12.16172 12.16175 12.16178 12.16181
#> 12 22.82941 22.82945 22.82948 22.82950 22.82952 22.82954 22.82956 22.82957
#> 13 17.19765 17.19764 17.19764 17.19763 17.19763 17.19763 17.19762 17.19762
#> 14 17.60570 17.60569 17.60569 17.60569 17.60569 17.60569 17.60569 17.60570
#> 15 14.01124 14.01124 14.01125 14.01125 14.01126 14.01126 14.01126 14.01126
#> 16 16.84274 16.84270 16.84266 16.84262 16.84259 16.84255 16.84252 16.84249
#> 17 25.40427 25.40425 25.40423 25.40422 25.40420 25.40419 25.40417 25.40416
#> 18 26.31096 26.31101 26.31105 26.31110 26.31114 26.31118 26.31121 26.31124
#> 19 16.76615 16.76613 16.76610 16.76608 16.76607 16.76605 16.76604 16.76603
#> 20 18.88415 18.88416 18.88417 18.88418 18.88418 18.88418 18.88418 18.88418
#> 21 22.53037 22.53040 22.53042 22.53044 22.53046 22.53048 22.53050 22.53051
#>      0.0078   0.0072   0.0067   0.0063   0.0059   0.0055   0.0051   0.0048
#> 1  21.30392 21.30389 21.32402 21.32405 21.32408 21.32411 21.32414 21.32417
#> 2  20.14127 20.14124 20.13372 20.13375 20.13379 20.13382 20.13386 20.13389
#> 3  17.26815 17.26813 17.28915 17.28923 17.28931 17.28939 17.28947 17.28954
#> 4  19.25647 19.25648 19.25113 19.25110 19.25107 19.25104 19.25101 19.25099
#> 5  14.07096 14.07093 14.06187 14.06201 14.06214 14.06227 14.06240 14.06252
#> 6  23.72539 23.72540 23.73116 23.73117 23.73117 23.73118 23.73118 23.73118
#> 7  19.41236 19.41237 19.40281 19.40269 19.40256 19.40244 19.40233 19.40221
#> 8  15.39974 15.39978 15.42535 15.42530 15.42526 15.42522 15.42517 15.42513
#> 9  15.31067 15.31072 15.37674 15.37677 15.37680 15.37683 15.37686 15.37689
#> 10 13.16645 13.16642 13.21218 13.21235 13.21251 13.21266 13.21282 13.21296
#> 11 12.16184 12.16186 12.14043 12.14035 12.14027 12.14020 12.14013 12.14006
#> 12 22.82958 22.82959 22.87441 22.87457 22.87473 22.87488 22.87503 22.87518
#> 13 17.19762 17.19762 17.11171 17.11151 17.11132 17.11113 17.11095 17.11078
#> 14 17.60571 17.60571 17.60683 17.60676 17.60669 17.60663 17.60656 17.60650
#> 15 14.01126 14.01126 13.94984 13.94973 13.94963 13.94954 13.94944 13.94935
#> 16 16.84246 16.84243 16.83281 16.83284 16.83286 16.83289 16.83291 16.83293
#> 17 25.40415 25.40414 25.41979 25.41980 25.41981 25.41983 25.41984 25.41985
#> 18 26.31128 26.31130 26.29467 26.29462 26.29456 26.29451 26.29446 26.29441
#> 19 16.76602 16.76602 16.80130 16.80138 16.80146 16.80154 16.80162 16.80169
#> 20 18.88418 18.88417 18.87148 18.87152 18.87157 18.87161 18.87165 18.87169
#> 21 22.53053 22.53054 22.48860 22.48849 22.48838 22.48827 22.48817 22.48807
#>      0.0044   0.0041   0.0039   0.0036
#> 1  21.32419 21.32421 21.32424 21.32426
#> 2  20.13392 20.13395 20.13398 20.13401
#> 3  17.28961 17.28968 17.28975 17.28981
#> 4  19.25096 19.25094 19.25091 19.25089
#> 5  14.06264 14.06275 14.06287 14.06297
#> 6  23.73119 23.73119 23.73119 23.73120
#> 7  19.40210 19.40200 19.40189 19.40179
#> 8  15.42510 15.42506 15.42502 15.42499
#> 9  15.37692 15.37694 15.37697 15.37699
#> 10 13.21311 13.21324 13.21338 13.21351
#> 11 12.13999 12.13993 12.13987 12.13981
#> 12 22.87532 22.87545 22.87558 22.87571
#> 13 17.11061 17.11044 17.11028 17.11013
#> 14 17.60644 17.60638 17.60632 17.60626
#> 15 13.94926 13.94918 13.94910 13.94902
#> 16 16.83295 16.83297 16.83299 16.83301
#> 17 25.41986 25.41988 25.41989 25.41990
#> 18 26.29437 26.29432 26.29428 26.29424
#> 19 16.80176 16.80183 16.80190 16.80196
#> 20 18.87173 18.87177 18.87180 18.87184
#> 21 22.48797 22.48788 22.48779 22.48770
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.4 18.7 18.1 14.3 24.4 19.2 17.3 15.2 10.4 14.7 21.5 15.5 15.2 13.3
#> [16] 19.2 27.3 26.0 15.8 19.7 21.4
#> 
#> $X
#>               am       carb        cyl        disp       drat       gear
#>  [1,]  1.5811388  0.9227055 -0.3464102 -0.86764912  0.8707440  0.6509446
#>  [2,] -0.6324555 -1.3130809 -0.3464102 -0.01299833 -0.7934304 -0.7160390
#>  [3,] -0.6324555 -0.5678188  0.8660254  0.87653616 -0.6513668 -0.7160390
#>  [4,] -0.6324555 -1.3130809 -0.3464102 -0.30078890 -1.4428643 -0.7160390
#>  [5,] -0.6324555  0.9227055  0.8660254  0.87653616 -0.5295979 -0.7160390
#>  [6,] -0.6324555 -0.5678188 -1.5588457 -0.98363744  0.4445530  0.6509446
#>  [7,] -0.6324555  0.9227055 -0.3464102 -0.80137008  0.9113336  0.6509446
#>  [8,] -0.6324555  0.1774434  0.8660254  0.14223416 -0.8137252 -0.7160390
#>  [9,] -0.6324555  0.1774434  0.8660254  0.14223416 -0.8137252 -0.7160390
#> [10,] -0.6324555  0.9227055  0.8660254  1.85327991 -1.0978526 -0.7160390
#> [11,] -0.6324555  0.9227055  0.8660254  1.57421027 -0.4890083 -0.7160390
#> [12,] -0.6324555 -1.3130809 -1.5588457 -1.21561408  0.4648478 -0.7160390
#> [13,] -0.6324555 -0.5678188  0.8660254  0.51025725 -1.4428643 -0.7160390
#> [14,] -0.6324555 -0.5678188  0.8660254  0.38816428 -0.6513668 -0.7160390
#> [15,] -0.6324555  0.9227055  0.8660254  0.78932689  0.5257322 -0.7160390
#> [16,] -0.6324555 -0.5678188  0.8660254  1.22537321 -0.7934304 -0.7160390
#> [17,]  1.5811388 -1.3130809 -1.5588457 -1.57404415  1.2360506  0.6509446
#> [18,]  1.5811388 -0.5678188 -1.5588457 -1.21386989  1.9463689  2.0179281
#> [19,]  1.5811388  0.9227055  0.8660254  0.79804782  1.5201779  2.0179281
#> [20,]  1.5811388  2.4132297 -0.3464102 -0.99846301  0.3024893  2.0179281
#> [21,]  1.5811388 -0.5678188 -1.5588457 -1.20776525  1.2969350  0.6509446
#>                hp        qsec         vs          wt
#>  [1,] -0.75978695 -0.71962188 -0.7071068 -0.92778847
#>  [2,] -0.75978695  1.19998329  1.4142136 -0.19960273
#>  [3,]  0.34705316 -0.35889071 -0.7071068  0.07576162
#>  [4,] -0.84492850  1.70243028  1.4142136  0.10023845
#>  [5,]  1.53903483 -1.11900282 -0.7071068  0.23486103
#>  [6,] -1.57714581  1.56071446  1.4142136 -0.23019877
#>  [7,] -0.53841893  0.46563769  1.4142136  0.07576162
#>  [8,]  0.43219471  0.01472372 -0.7071068  0.43067568
#>  [9,]  0.43219471  0.27238884 -0.7071068  0.49186776
#> [10,]  0.85790245  0.25950559 -0.7071068  2.29091488
#> [11,]  1.28361019 -0.10122558 -0.7071068  2.40717983
#> [12,] -0.98115498  1.56715609  1.4142136 -1.11748392
#> [13,] -0.07865457 -0.45551513 -0.7071068  0.17366895
#> [14,] -0.07865457 -0.17852512 -0.7071068  0.06964241
#> [15,]  1.53903483 -1.39599283 -0.7071068  0.56529825
#> [16,]  0.34705316 -0.33956582 -0.7071068  0.57141746
#> [17,] -1.50903257  0.85213537  1.4142136 -1.76611996
#> [18,] -1.08332483 -0.56502281 -0.7071068 -1.51523243
#> [19,]  1.86257271 -1.98218099 -0.7071068 -0.25467561
#> [20,]  0.34705316 -1.33801818 -0.7071068 -0.74421224
#> [21,] -0.77681526  0.65888653  1.4142136 -0.73197382
#> attr(,"center")
#>  [1]   0.2857143   2.7619048   6.5714286 259.4904762   3.4709524   3.5238095
#>  [7] 154.6190476  17.5771429   0.3333333   3.3780952
#> attr(,"scale")
#>  [1]   0.4517540   1.3418098   1.6495722 114.6667175   0.4927368   0.7315377
#>  [7]  58.7257356   1.5524026   0.4714045   0.8170992
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 31
#> 
#> $lambda.min
#> [1] 0.4435434
#> 
#> $null.dev
#> [1] 17.49392
#> 
#> $Bias
#> [1] 4.705734
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 16.52341