Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 31.568957 30.223050 26.673489 23.471286 20.805700 18.598781 16.783274
#>   [8] 15.301190 14.102567 13.144390 12.389657 11.806569 11.367827 11.050020
#>  [15] 10.833096 10.674226 10.363346 10.081042 10.101586 10.129318 10.122104
#>  [22] 10.011835  9.874910  9.758039  9.664686  9.590147  9.479818  9.392488
#>  [29]  9.328932  9.474125  9.269339  9.083725  8.690637  8.214554  7.684343
#>  [36]  7.514980  7.369927  6.489539  6.457788  6.475507  6.992874  7.311004
#>  [43]  7.237472  7.227427  7.146636  7.102419  7.498396  7.400832  7.394047
#>  [50]  7.398284  7.523122  8.440021  8.492293  8.901227  8.901019  8.900818
#>  [57]  9.100252  8.755900 12.095373 17.897793 18.468446 18.569253 19.699460
#>  [64] 19.729440 19.807023 19.884296 19.884963 19.878096 19.863215 19.864163
#>  [71] 19.865107 19.866047 19.866983 19.857531 19.886773 20.479152 20.480761
#>  [78] 20.560748 20.562350 20.646872 20.730946 20.693700 20.694540 20.695389
#>  [85] 20.696241 20.697094 20.697946 20.698792 20.699633 20.700466 20.701289
#>  [92] 20.702102 20.702904 20.703694 20.712780 20.713580 20.709248 20.710152
#>  [99] 20.711020 20.730550
#> 
#> $cvse
#>   [1] 9.758419 9.780197 8.923635 8.104563 7.375390 6.724942 6.143921 5.624622
#>   [9] 5.160678 4.746849 4.378815 4.053010 3.766449 3.516581 3.301143 3.115874
#>  [17] 2.944173 2.837969 2.839614 2.841843 2.846481 2.853559 2.863594 2.872123
#>  [25] 2.878780 2.882925 2.859591 2.821815 2.789091 2.796287 2.856464 2.868472
#>  [33] 2.854739 2.844380 2.720080 2.624519 2.539630 2.505611 2.480841 2.474646
#>  [41] 2.475599 2.476515 2.437104 2.393380 2.360875 2.331155 2.638723 2.522334
#>  [49] 2.490205 2.490756 2.526638 2.874632 2.871763 2.867232 2.867232 2.867229
#>  [57] 2.836927 2.784795 3.298572 5.774906 5.749609 5.741562 5.778223 5.777000
#>  [65] 5.760898 5.782566 5.783237 5.785159 5.783189 5.783811 5.784420 5.785016
#>  [73] 5.785601 5.796912 5.810723 6.305273 6.306191 6.317257 6.318136 6.318686
#>  [81] 6.319340 6.305182 6.305861 6.306531 6.307191 6.307840 6.308477 6.309102
#>  [89] 6.309714 6.310314 6.310900 6.311474 6.312035 6.312584 6.313445 6.313968
#>  [97] 6.314855 6.315342 6.315820 6.315785
#> 
#> $fold
#>  [1]  3  9  9  8  5 10  8 10  1  7  1  5  2  1  3  6  4  4  6  7  2
#> 
#> $lambda
#>   [1] 4.703452157 4.386455223 4.090822822 3.815115056 3.557989070 3.318192514
#>   [7] 3.094557445 2.885994630 2.691488251 2.510090951 2.340919222 2.183149101
#>  [13] 2.036012158 1.898791753 1.770819544 1.651472234 1.540168532 1.436366328
#>  [19] 1.339560045 1.249278181 1.165081013 1.086558452 1.013328049 0.945033130
#>  [25] 0.881341060 0.821941622 0.766545508 0.714882906 0.666702191 0.621768695
#>  [31] 0.579863566 0.540782702 0.504335758 0.470345216 0.438645522 0.409082282
#>  [37] 0.381511506 0.355798907 0.331819251 0.309455744 0.288599463 0.269148825
#>  [43] 0.251009095 0.234091922 0.218314910 0.203601216 0.189879175 0.177081954
#>  [49] 0.165147223 0.154016853 0.143636633 0.133956004 0.124927818 0.116508101
#>  [55] 0.108655845 0.101332805 0.094503313 0.088134106 0.082194162 0.076654551
#>  [61] 0.071488291 0.066670219 0.062176869 0.057986357 0.054078270 0.050433576
#>  [67] 0.047034522 0.043864552 0.040908228 0.038151151 0.035579891 0.033181925
#>  [73] 0.030945574 0.028859946 0.026914883 0.025100910 0.023409192 0.021831491
#>  [79] 0.020360122 0.018987918 0.017708195 0.016514722 0.015401685 0.014363663
#>  [85] 0.013395600 0.012492782 0.011650810 0.010865585 0.010133280 0.009450331
#>  [91] 0.008813411 0.008219416 0.007665455 0.007148829 0.006667022 0.006217687
#>  [97] 0.005798636 0.005407827 0.005043358 0.004703452
#> 
#> $fit
#> $beta
#>              4.7035     4.3865     4.0908     3.8151    3.5580    3.3182
#> (Intercept) 19.6619 21.1044635 22.7337977 24.2533202 25.670432 26.992035
#> am           0.0000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> carb         0.0000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> cyl          0.0000 -0.2404265 -0.5119822 -0.7652359 -1.001421 -1.221688
#> disp         0.0000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> drat         0.0000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> gear         0.0000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> hp           0.0000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> qsec         0.0000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> vs           0.0000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> wt           0.0000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#>               3.0946    2.8860    2.6915    2.5101    2.3409    2.1831
#> (Intercept) 28.22457 29.374030 30.446023 31.445767 32.378132 33.247659
#> am           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.42711 -1.618688 -1.797353 -1.963977 -2.119371 -2.264292
#> disp         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                2.0360    1.8988    1.7708    1.6515    1.5402    1.4364
#> (Intercept) 34.058582 34.814852 35.520152 36.177917 36.740476 36.740476
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.399446 -2.525491 -2.643041 -2.752669 -2.846429 -2.846429
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.3396    1.2493    1.1651     1.0866     1.0133     0.9450
#> (Intercept) 36.740476 36.740476 36.590684 36.3681334 36.1605783 35.9670118
#> am           0.000000  0.000000  0.174758  0.4343999  0.6765475  0.9023752
#> carb         0.000000  0.000000  0.000000  0.0000000  0.0000000  0.0000000
#> cyl         -2.846429 -2.846429 -2.833946 -2.8154000 -2.7981037 -2.7819732
#> disp         0.000000  0.000000  0.000000  0.0000000  0.0000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.0000000  0.0000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.0000000  0.0000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.0000000  0.0000000  0.0000000
#> qsec         0.000000  0.000000  0.000000  0.0000000  0.0000000  0.0000000
#> vs           0.000000  0.000000  0.000000  0.0000000  0.0000000  0.0000000
#> wt           0.000000  0.000000  0.000000  0.0000000  0.0000000  0.0000000
#>                0.8813    0.8219    0.7665    0.7149    0.6667    0.6218
#> (Intercept) 35.786491 35.618137 35.461129 39.013253 39.013096 39.013010
#> am           1.112983  1.309396  1.492572  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.766930 -2.752900 -2.739816 -1.785404 -1.785303 -1.785247
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000 -2.679369 -2.679509 -2.679585
#>                0.5799    0.5408    0.5043    0.4703    0.4386    0.4091
#> (Intercept) 39.012963 39.012937 39.012923 39.012916 39.012912 39.012909
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.785217 -1.785201 -1.785191 -1.785187 -1.785184 -1.785182
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.679627 -2.679650 -2.679662 -2.679669 -2.679673 -2.679675
#>                0.3815    0.3558    0.3318    0.3095    0.2886    0.2691
#> (Intercept) 39.012908 39.012907 39.012907 39.012907 39.012907 39.012907
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.785182 -1.785181 -1.785181 -1.785181 -1.785181 -1.785181
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.679676 -2.679677 -2.679677 -2.679677 -2.679677 -2.679677
#>                  0.2510      0.2341      0.2183     0.2036      0.1899
#> (Intercept) 35.59297460 35.59270090 35.59250502 35.5923653 32.79811153
#> am           1.65076206  1.65071102  1.65066424  1.6506231  2.28825438
#> carb         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> cyl         -1.23153565 -1.23131413 -1.23114570 -1.2310180 -0.94714154
#> disp         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> drat         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> hp          -0.01909879 -0.01910117 -0.01910291 -0.0191042 -0.01985427
#> qsec         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.0000000  1.19041937
#> wt          -2.01682898 -2.01704379 -2.01721230 -2.0173438 -1.90608473
#>                  0.1771      0.1651      0.1540      0.1436      0.1340
#> (Intercept) 32.79691449 32.79598083 32.79525140 32.79468083 32.79423413
#> am           2.28849182  2.28867367  2.28881338  2.28892101  2.28900410
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl         -0.94689830 -0.94670565 -0.94655312 -0.94643243 -0.94633697
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.01985576 -0.01985695 -0.01985788 -0.01985861 -0.01985919
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           1.19081055  1.19111608  1.19135513  1.19154241  1.19168925
#> wt          -1.90618885 -1.90627439 -1.90634427 -1.90640108 -1.90644708
#>                  0.1249      0.1165      0.1087      0.1013      0.0945
#> (Intercept) 32.79388419 32.79360994 32.79339495 32.79322639 32.79586525
#> am           2.28906836  2.28911814  2.28915674  2.28918670  2.33509962
#> carb         0.00000000  0.00000000  0.00000000  0.00000000 -0.05153441
#> cyl         -0.94626151 -0.94620191 -0.94615485 -0.94611773 -0.95663438
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.01985965 -0.01986001 -0.01986029 -0.01986051 -0.01891985
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           1.19180447  1.19189491  1.19196591  1.19202166  1.16526640
#> wt          -1.90648420 -1.90651407 -1.90653803 -1.90655722 -1.88396221
#>                  0.0881      0.0822      0.0767      0.0715      0.0667
#> (Intercept) 30.31177427 21.61555343 21.61129565 21.60780453 21.60496195
#> am           2.08927474  2.43610307  2.43624778  2.43636601  2.43646193
#> carb        -0.28337345 -0.22661182 -0.22671886 -0.22681097 -0.22688999
#> cyl         -0.84385445 -0.53767120 -0.53755537 -0.53746835 -0.53740493
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.56615764  0.67578587  0.67593299  0.67605355  0.67615157
#> hp          -0.01842055 -0.01714384 -0.01713952 -0.01713576 -0.01713249
#> qsec         0.00000000  0.41355430  0.41373868  0.41389143  0.41401726
#> vs           1.08311885  0.56229012  0.56205461  0.56185015  0.56167306
#> wt          -1.75406941 -2.14313556 -2.14331086 -2.14345259 -2.14356613
#>                  0.0622      0.0580      0.0541        0.0504       0.0470
#> (Intercept) 21.60266597 21.60082886 21.59937532  2.159774e+01 21.482903210
#> am           2.43653911  2.43660062  2.43664907  2.436568e+00  2.437844917
#> carb        -0.22695756 -0.22701515 -0.22706406 -2.270904e-01 -0.601905574
#> cyl         -0.53736068 -0.53733189 -0.53731544 -5.372537e-01 -0.470657734
#> disp         0.00000000  0.00000000  0.00000000 -3.093826e-06 -0.012839251
#> drat         0.00000000  0.00000000  0.00000000  0.000000e+00  0.000000000
#> gear         0.67623056  0.67629355  0.67634314  6.762750e-01  0.703224576
#> hp          -0.01712965 -0.01712719 -0.01712507 -1.711978e-02 -0.006863643
#> qsec         0.41412030  0.41420406  0.41427161  4.143700e-01  0.346565544
#> vs           0.56152007  0.56138823  0.56127490  5.610343e-01  0.000000000
#> wt          -2.14365612 -2.14372651 -2.14378067 -2.143812e+00 -1.039402106
#>                   0.0439       0.0409       0.0382       0.0356       0.0332
#> (Intercept) 21.483165085 21.483417785 21.476463493 20.282308566 20.278794314
#> am           2.437866600  2.437887318  2.436866981  2.468380024  2.468456604
#> carb        -0.602053867 -0.602196328 -0.602385619 -0.603761903 -0.603836187
#> cyl         -0.470604893 -0.470554030 -0.470292957 -0.449519810 -0.449256046
#> disp        -0.012843643 -0.012847866 -0.012855430 -0.013304836 -0.013305351
#> drat         0.000000000  0.000000000  0.002612757  0.376875918  0.376806769
#> gear         0.703290684  0.703354263  0.702789599  0.534809445  0.535092063
#> hp          -0.006861186 -0.006858829 -0.006853136 -0.006144683 -0.006145647
#> qsec         0.346494840  0.346426812  0.346279372  0.351063706  0.351162032
#> vs           0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> wt          -1.038941232 -1.038498106 -1.037696779 -0.958110353 -0.958179380
#>                   0.0309       0.0289       0.0269       0.0251       0.0234
#> (Intercept) 20.757692508 20.758053326 20.758394262 20.758717347 20.759024323
#> am           2.530645865  2.530621424  2.530598143  2.530575931  2.530554709
#> carb        -0.581798388 -0.581656760 -0.581520617 -0.581389734 -0.581263899
#> cyl         -0.437359719 -0.437421815 -0.437481409 -0.437538648 -0.437593665
#> disp        -0.012612679 -0.012608408 -0.012604297 -0.012600339 -0.012596531
#> drat         0.388889127  0.388795178  0.388704282  0.388616372  0.388531382
#> gear         0.513948422  0.513919833  0.513892706  0.513866919  0.513842362
#> hp          -0.007101363 -0.007104004 -0.007106538 -0.007108969 -0.007111302
#> qsec         0.309287231  0.309335523  0.309382364  0.309427750  0.309471682
#> vs           0.287189016  0.287242404  0.287293286  0.287341802  0.287388084
#> wt          -0.959351952 -0.959787452 -0.960207101 -0.960611400 -0.961000843
#>                   0.0218      0.0204       0.0190       0.0177       0.0165
#> (Intercept) 20.759316689 20.75959573 20.759862563 20.760118139 20.760363288
#> am           2.530534406  2.53051496  2.530496324  2.530478444  2.530461279
#> carb        -0.581142909 -0.58102657 -0.580914700 -0.580807121 -0.580703665
#> cyl         -0.437646579 -0.43769750 -0.437746517 -0.437793726 -0.437839206
#> disp        -0.012592866 -0.01258934 -0.012585945 -0.012582679 -0.012579537
#> drat         0.388449246  0.38836990  0.388293275  0.388219311  0.388147940
#> gear         0.513818939  0.51379656  0.513775149  0.513754635  0.513734955
#> hp          -0.007113541 -0.00711569 -0.007117753 -0.007119734 -0.007121636
#> qsec         0.309514172  0.30955524  0.309594893  0.309633171  0.309670097
#> vs           0.287432252  0.28747442  0.287514698  0.287553180  0.287589960
#> wt          -0.961375914 -0.96173709 -0.962084829 -0.962419592 -0.962741820
#>                   0.0154       0.0144       0.0134       0.0125       0.0117
#> (Intercept) 20.760598727 20.760825081 20.761042894 20.761252646 20.761454757
#> am           2.530444791  2.530428947  2.530413714  2.530399065  2.530384973
#> carb        -0.580604171 -0.580508487 -0.580416464 -0.580327963 -0.580242847
#> cyl         -0.437883032 -0.437925271 -0.437965987 -0.438005239 -0.438043082
#> disp        -0.012576514 -0.012573606 -0.012570808 -0.012568117 -0.012565529
#> drat         0.388079095  0.388012710  0.387948718  0.387887051  0.387827643
#> gear         0.513716055  0.513697885  0.513680400  0.513663562  0.513647336
#> hp          -0.007123462 -0.007125215 -0.007126899 -0.007128516 -0.007130069
#> qsec         0.309705701  0.309740017  0.309773078  0.309804920  0.309835577
#> vs           0.287625127  0.287658762  0.287690941  0.287721737  0.287751217
#> wt          -0.963051947 -0.963350393 -0.963637568 -0.963913872 -0.964179691
#>                  0.0109       0.0101      0.0095       0.0088       0.0082
#> (Intercept) 20.76164960 20.761837515 20.76201880 20.762193717 20.762362526
#> am           2.53037141  2.530358368  2.53034581  2.530333727  2.530322095
#> carb        -0.58016099 -0.580082262 -0.58000655 -0.579933733 -0.579863706
#> cyl         -0.43807957 -0.438114747 -0.43814866 -0.438181362 -0.438212885
#> disp        -0.01256304 -0.012560644 -0.01255834 -0.012556126 -0.012553997
#> drat         0.38777043  0.387715335  0.38766230  0.387611259  0.387562146
#> gear         0.51363169  0.513616595  0.51360203  0.513587962  0.513574380
#> hp          -0.00713156 -0.007132993 -0.00713437 -0.007135692 -0.007136962
#> qsec         0.30986509  0.309893486  0.30992081  0.309947093  0.309972373
#> vs           0.28777945  0.287806482  0.28783238  0.287857203  0.287880991
#> wt          -0.96443540 -0.964681372 -0.96491795 -0.965145486 -0.965364306
#>                   0.0077       0.0071      0.0067       0.0062       0.0058
#> (Intercept) 20.762525451 20.762682703 20.76283448 20.762980963 20.763122331
#> am           2.530310898  2.530300121  2.53028975  2.530279763  2.530270152
#> carb        -0.579796362 -0.579731599 -0.57966932 -0.579609429 -0.579551837
#> cyl         -0.438243273 -0.438272563 -0.43830079 -0.438328000 -0.438354216
#> disp        -0.012551948 -0.012549979 -0.01254809 -0.012546264 -0.012544514
#> drat         0.387514895  0.387469446  0.38742573  0.387383701  0.387343287
#> gear         0.513561262  0.513548589  0.51353635  0.513524518  0.513513091
#> hp          -0.007138182 -0.007139354 -0.00714048 -0.007141562 -0.007142601
#> qsec         0.309996684  0.310020059  0.31004253  0.310064138  0.310084906
#> vs           0.287903795  0.287925660  0.28794663  0.287966742  0.287986037
#> wt          -0.965574734 -0.965777081 -0.96597165 -0.966158722 -0.966338588
#>                  0.0054       0.0050       0.0047
#> (Intercept) 20.76325875 20.763390372 20.763517357
#> am           2.53026090  2.530252001  2.530243434
#> carb        -0.57949646 -0.579443206 -0.579392001
#> cyl         -0.43837948 -0.438403811 -0.438427253
#> disp        -0.01254283 -0.012541212 -0.012539656
#> drat         0.38730443  0.387267083  0.387231183
#> gear         0.51350205  0.513491388  0.513481088
#> hp          -0.00714360 -0.007144559 -0.007145481
#> qsec         0.31010487  0.310124053  0.310142493
#> vs           0.28800455  0.288022314  0.288039364
#> wt          -0.96651151 -0.966677766 -0.966837592
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   5   4   4   4   4   4   4  38   1   1   1   1   1   1   1   1
#>  [37]   1   1   1   1   1   1  40   1   1   1 142   1   1   1   1   1   1   1
#>  [55]   1   1 131  75  47   1   1   1   1   1   1   2 161   1   1   2  42   1
#>  [73]  63   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.703452157 4.386455223 4.090822822 3.815115056 3.557989070 3.318192514
#>   [7] 3.094557445 2.885994630 2.691488251 2.510090951 2.340919222 2.183149101
#>  [13] 2.036012158 1.898791753 1.770819544 1.651472234 1.540168532 1.436366328
#>  [19] 1.339560045 1.249278181 1.165081013 1.086558452 1.013328049 0.945033130
#>  [25] 0.881341060 0.821941622 0.766545508 0.714882906 0.666702191 0.621768695
#>  [31] 0.579863566 0.540782702 0.504335758 0.470345216 0.438645522 0.409082282
#>  [37] 0.381511506 0.355798907 0.331819251 0.309455744 0.288599463 0.269148825
#>  [43] 0.251009095 0.234091922 0.218314910 0.203601216 0.189879175 0.177081954
#>  [49] 0.165147223 0.154016853 0.143636633 0.133956004 0.124927818 0.116508101
#>  [55] 0.108655845 0.101332805 0.094503313 0.088134106 0.082194162 0.076654551
#>  [61] 0.071488291 0.066670219 0.062176869 0.057986357 0.054078270 0.050433576
#>  [67] 0.047034522 0.043864552 0.040908228 0.038151151 0.035579891 0.033181925
#>  [73] 0.030945574 0.028859946 0.026914883 0.025100910 0.023409192 0.021831491
#>  [79] 0.020360122 0.018987918 0.017708195 0.016514722 0.015401685 0.014363663
#>  [85] 0.013395600 0.012492782 0.011650810 0.010865585 0.010133280 0.009450331
#>  [91] 0.008813411 0.008219416 0.007665455 0.007148829 0.006667022 0.006217687
#>  [97] 0.005798636 0.005407827 0.005043358 0.004703452
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 27
#> 
#> $loss
#>   [1] 586.52952 513.11864 437.98865 375.36513 323.43573 280.63652 245.61881
#>   [8] 217.22030 194.44014 176.41716 162.41110 151.78623 143.99723 138.57685
#>  [15] 135.12533 133.30110 132.80881 132.80881 132.80881 132.80881 128.53328
#>  [22] 122.72885 117.90569 113.92091 110.65114 107.99004 105.84602  80.45537
#>  [29]  80.45537  80.45537  80.45537  80.45537  80.45537  80.45537  80.45537
#>  [36]  80.45537  80.45537  80.45537  80.45537  80.45537  80.45537  80.45537
#>  [43]  68.94570  68.94570  68.94570  68.94570  66.55861  66.55861  66.55861
#>  [50]  66.55860  66.55860  66.55860  66.55860  66.55860  66.55860  66.55860
#>  [57]  66.24512  65.20849  63.70408  63.70408  63.70407  63.70407  63.70407
#>  [64]  63.70406  63.70406  63.70331  62.30243  62.30242  62.30241  62.30043
#>  [71]  62.16118  62.16117  62.06293  62.06292  62.06291  62.06290  62.06289
#>  [78]  62.06288  62.06288  62.06287  62.06286  62.06286  62.06285  62.06285
#>  [85]  62.06284  62.06284  62.06284  62.06283  62.06283  62.06282  62.06282
#>  [92]  62.06282  62.06282  62.06281  62.06281  62.06281  62.06281  62.06281
#>  [99]  62.06281  62.06280
#> 
#> $linear.predictors
#>     4.7035   4.3865   4.0908   3.8151   3.5580   3.3182   3.0946   2.8860
#> 1  19.6619 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190
#> 2  19.6619 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190
#> 3  19.6619 20.14276 20.68587 21.19238 21.66475 22.10528 22.51613 22.89928
#> 4  19.6619 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190
#> 5  19.6619 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190
#> 6  19.6619 19.18105 18.63794 18.13143 17.65906 17.21853 16.80768 16.42453
#> 7  19.6619 20.14276 20.68587 21.19238 21.66475 22.10528 22.51613 22.89928
#> 8  19.6619 20.14276 20.68587 21.19238 21.66475 22.10528 22.51613 22.89928
#> 9  19.6619 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190
#> 10 19.6619 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190
#> 11 19.6619 19.18105 18.63794 18.13143 17.65906 17.21853 16.80768 16.42453
#> 12 19.6619 19.18105 18.63794 18.13143 17.65906 17.21853 16.80768 16.42453
#> 13 19.6619 19.18105 18.63794 18.13143 17.65906 17.21853 16.80768 16.42453
#> 14 19.6619 20.14276 20.68587 21.19238 21.66475 22.10528 22.51613 22.89928
#> 15 19.6619 20.14276 20.68587 21.19238 21.66475 22.10528 22.51613 22.89928
#> 16 19.6619 19.18105 18.63794 18.13143 17.65906 17.21853 16.80768 16.42453
#> 17 19.6619 20.14276 20.68587 21.19238 21.66475 22.10528 22.51613 22.89928
#> 18 19.6619 20.14276 20.68587 21.19238 21.66475 22.10528 22.51613 22.89928
#> 19 19.6619 19.18105 18.63794 18.13143 17.65906 17.21853 16.80768 16.42453
#> 20 19.6619 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190
#> 21 19.6619 19.18105 18.63794 18.13143 17.65906 17.21853 16.80768 16.42453
#>      2.6915   2.5101   2.3409   2.1831   2.0360   1.8988   1.7708   1.6515
#> 1  19.66190 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190
#> 2  19.66190 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190
#> 3  23.25661 23.58986 23.90065 24.19049 24.46080 24.71289 24.94799 25.16724
#> 4  19.66190 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190
#> 5  19.66190 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190
#> 6  16.06720 15.73395 15.42316 15.13332 14.86301 14.61092 14.37582 14.15657
#> 7  23.25661 23.58986 23.90065 24.19049 24.46080 24.71289 24.94799 25.16724
#> 8  23.25661 23.58986 23.90065 24.19049 24.46080 24.71289 24.94799 25.16724
#> 9  19.66190 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190
#> 10 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190
#> 11 16.06720 15.73395 15.42316 15.13332 14.86301 14.61092 14.37582 14.15657
#> 12 16.06720 15.73395 15.42316 15.13332 14.86301 14.61092 14.37582 14.15657
#> 13 16.06720 15.73395 15.42316 15.13332 14.86301 14.61092 14.37582 14.15657
#> 14 23.25661 23.58986 23.90065 24.19049 24.46080 24.71289 24.94799 25.16724
#> 15 23.25661 23.58986 23.90065 24.19049 24.46080 24.71289 24.94799 25.16724
#> 16 16.06720 15.73395 15.42316 15.13332 14.86301 14.61092 14.37582 14.15657
#> 17 23.25661 23.58986 23.90065 24.19049 24.46080 24.71289 24.94799 25.16724
#> 18 23.25661 23.58986 23.90065 24.19049 24.46080 24.71289 24.94799 25.16724
#> 19 16.06720 15.73395 15.42316 15.13332 14.86301 14.61092 14.37582 14.15657
#> 20 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190 19.66190
#> 21 16.06720 15.73395 15.42316 15.13332 14.86301 14.61092 14.37582 14.15657
#>      1.5402   1.4364   1.3396   1.2493   1.1651   1.0866   1.0133   0.9450
#> 1  19.66190 19.66190 19.66190 19.66190 19.76177 19.91013 20.04850 20.17755
#> 2  19.66190 19.66190 19.66190 19.66190 19.76177 19.91013 20.04850 20.17755
#> 3  25.35476 25.35476 25.35476 25.35476 25.42966 25.54093 25.64471 25.74149
#> 4  19.66190 19.66190 19.66190 19.66190 19.58701 19.47573 19.37196 19.27517
#> 5  19.66190 19.66190 19.66190 19.66190 19.58701 19.47573 19.37196 19.27517
#> 6  13.96905 13.96905 13.96905 13.96905 13.91912 13.84493 13.77575 13.71123
#> 7  25.35476 25.35476 25.35476 25.35476 25.25490 25.10653 24.96816 24.83912
#> 8  25.35476 25.35476 25.35476 25.35476 25.25490 25.10653 24.96816 24.83912
#> 9  19.66190 19.66190 19.66190 19.66190 19.58701 19.47573 19.37196 19.27517
#> 10 19.66190 19.66190 19.66190 19.66190 19.58701 19.47573 19.37196 19.27517
#> 11 13.96905 13.96905 13.96905 13.96905 13.91912 13.84493 13.77575 13.71123
#> 12 13.96905 13.96905 13.96905 13.96905 13.91912 13.84493 13.77575 13.71123
#> 13 13.96905 13.96905 13.96905 13.96905 13.91912 13.84493 13.77575 13.71123
#> 14 25.35476 25.35476 25.35476 25.35476 25.42966 25.54093 25.64471 25.74149
#> 15 25.35476 25.35476 25.35476 25.35476 25.25490 25.10653 24.96816 24.83912
#> 16 13.96905 13.96905 13.96905 13.96905 13.91912 13.84493 13.77575 13.71123
#> 17 25.35476 25.35476 25.35476 25.35476 25.42966 25.54093 25.64471 25.74149
#> 18 25.35476 25.35476 25.35476 25.35476 25.42966 25.54093 25.64471 25.74149
#> 19 13.96905 13.96905 13.96905 13.96905 14.09387 14.27933 14.45230 14.61360
#> 20 19.66190 19.66190 19.66190 19.66190 19.76177 19.91013 20.04850 20.17755
#> 21 13.96905 13.96905 13.96905 13.96905 14.09387 14.27933 14.45230 14.61360
#>      0.8813   0.8219   0.7665   0.7149   0.6667   0.6218   0.5799   0.5408
#> 1  20.29789 20.41013 20.51480 21.28088 21.28097 21.28101 21.28104 21.28105
#> 2  20.29789 20.41013 20.51480 20.59764 20.59769 20.59772 20.59773 20.59774
#> 3  25.83175 25.91593 25.99444 25.65550 25.65542 25.65538 25.65536 25.65535
#> 4  19.18491 19.10073 19.02223 19.68666 19.68666 19.68666 19.68666 19.68666
#> 5  19.18491 19.10073 19.02223 19.03021 19.03018 19.03016 19.03015 19.03015
#> 6  13.65105 13.59493 13.54260 15.16467 15.16483 15.16491 15.16496 15.16498
#> 7  24.71877 24.60654 24.50186 23.32445 23.32425 23.32414 23.32408 23.32405
#> 8  24.71877 24.60654 24.50186 23.43162 23.43143 23.43133 23.43127 23.43124
#> 9  19.18491 19.10073 19.02223 19.08380 19.08377 19.08375 19.08374 19.08374
#> 10 19.18491 19.10073 19.02223 19.08380 19.08377 19.08375 19.08374 19.08374
#> 11 13.65105 13.59493 13.54260 13.82499 13.82507 13.82512 13.82514 13.82516
#> 12 13.65105 13.59493 13.54260 10.66333 10.66325 10.66321 10.66318 10.66317
#> 13 13.65105 13.59493 13.54260 10.19712 10.19702 10.19696 10.19693 10.19691
#> 14 25.83175 25.91593 25.99444 25.97702 25.97697 25.97693 25.97692 25.97691
#> 15 24.71877 24.60654 24.50186 25.26699 25.26690 25.26684 25.26681 25.26680
#> 16 13.65105 13.59493 13.54260 15.52639 15.52656 15.52666 15.52671 15.52674
#> 17 25.83175 25.91593 25.99444 26.68706 26.68704 26.68702 26.68702 26.68701
#> 18 25.83175 25.91593 25.99444 26.13779 26.13774 26.13771 26.13769 26.13768
#> 19 14.76404 14.90433 15.03517 16.23642 16.23663 16.23675 16.23681 16.23684
#> 20 20.29789 20.41013 20.51480 20.87898 20.87904 20.87907 20.87909 20.87910
#> 21 14.76404 14.90433 15.03517 15.16467 15.16483 15.16491 15.16496 15.16498
#>      0.5043   0.4703   0.4386   0.4091   0.3815   0.3558   0.3318   0.3095
#> 1  21.28106 21.28106 21.28107 21.28107 21.28107 21.28107 21.28107 21.28107
#> 2  20.59774 20.59775 20.59775 20.59775 20.59775 20.59775 20.59775 20.59775
#> 3  25.65534 25.65534 25.65533 25.65533 25.65533 25.65533 25.65533 25.65533
#> 4  19.68666 19.68666 19.68666 19.68666 19.68666 19.68666 19.68666 19.68666
#> 5  19.03014 19.03014 19.03014 19.03014 19.03014 19.03014 19.03014 19.03014
#> 6  15.16500 15.16500 15.16501 15.16501 15.16501 15.16501 15.16501 15.16501
#> 7  23.32403 23.32402 23.32402 23.32402 23.32401 23.32401 23.32401 23.32401
#> 8  23.43122 23.43121 23.43121 23.43120 23.43120 23.43120 23.43120 23.43120
#> 9  19.08374 19.08373 19.08373 19.08373 19.08373 19.08373 19.08373 19.08373
#> 10 19.08374 19.08373 19.08373 19.08373 19.08373 19.08373 19.08373 19.08373
#> 11 13.82517 13.82517 13.82517 13.82517 13.82517 13.82517 13.82517 13.82517
#> 12 10.66316 10.66316 10.66316 10.66316 10.66316 10.66316 10.66316 10.66316
#> 13 10.19690 10.19690 10.19689 10.19689 10.19689 10.19689 10.19689 10.19689
#> 14 25.97690 25.97690 25.97690 25.97689 25.97689 25.97689 25.97689 25.97689
#> 15 25.26679 25.26678 25.26678 25.26678 25.26678 25.26678 25.26678 25.26678
#> 16 15.52675 15.52676 15.52676 15.52677 15.52677 15.52677 15.52677 15.52677
#> 17 26.68701 26.68701 26.68701 26.68701 26.68701 26.68701 26.68701 26.68701
#> 18 26.13768 26.13768 26.13768 26.13768 26.13767 26.13767 26.13767 26.13767
#> 19 16.23686 16.23687 16.23688 16.23688 16.23688 16.23688 16.23688 16.23688
#> 20 20.87911 20.87911 20.87911 20.87912 20.87912 20.87912 20.87912 20.87912
#> 21 15.16500 15.16500 15.16501 15.16501 15.16501 15.16501 15.16501 15.16501
#>      0.2886   0.2691   0.2510   0.2341   0.2183   0.2036   0.1899   0.1771
#> 1  21.28107 21.28107 22.46956 22.46974 22.46988 22.46998 22.22560 22.22567
#> 2  20.59775 20.59775 21.95527 21.95540 21.95549 21.95556 21.73955 21.73959
#> 3  25.65533 25.65533 25.86236 25.86221 25.86208 25.86199 26.21966 26.21968
#> 4  19.68666 19.68666 19.61879 19.61889 19.61897 19.61904 19.99365 19.99380
#> 5  19.03014 19.03014 19.22016 19.22022 19.22027 19.22031 19.62593 19.62607
#> 6  15.16501 15.16501 13.86141 13.86156 13.86168 13.86178 13.55196 13.55197
#> 7  23.32401 23.32401 23.04902 23.04880 23.04863 23.04851 22.88859 22.88833
#> 8  23.43120 23.43120 22.49944 22.49915 22.49893 22.49876 22.30964 22.30934
#> 9  19.08373 19.08373 18.91672 18.91674 18.91676 18.91678 19.30667 19.30679
#> 10 19.08373 19.08373 18.91672 18.91674 18.91676 18.91678 19.30667 19.30679
#> 11 13.82517 13.82517 14.09441 14.09461 14.09476 14.09488 13.88945 13.88950
#> 12 10.66316 10.66316 11.23708 11.23697 11.23688 11.23681 11.14391 11.14380
#> 13 10.19689 10.19689 10.69517 10.69499 10.69485 10.69475 10.61371 10.61357
#> 14 25.97689 25.97689 26.62005 26.61998 26.61993 26.61988 26.98445 26.98453
#> 15 25.26678 25.26678 23.84277 23.84262 23.84251 23.84243 23.57560 23.57537
#> 16 15.52677 15.52677 15.94806 15.94847 15.94878 15.94902 15.69544 15.69560
#> 17 26.68701 26.68701 27.15451 27.15450 27.15449 27.15448 27.48956 27.48967
#> 18 26.13767 26.13767 26.26359 26.26348 26.26339 26.26332 25.41204 25.41169
#> 19 16.23688 16.23688 15.95602 15.95616 15.95627 15.95636 16.22542 16.22568
#> 20 20.87912 20.87912 20.92562 20.92561 20.92561 20.92560 20.64916 20.64911
#> 21 15.16501 15.16501 13.79328 13.79316 13.79308 13.79302 14.05333 14.05344
#>      0.1651   0.1540   0.1436   0.1340   0.1249   0.1165   0.1087   0.1013
#> 1  22.22572 22.22576 22.22579 22.22581 22.22583 22.22585 22.22586 22.22587
#> 2  21.73962 21.73964 21.73966 21.73967 21.73968 21.73969 21.73969 21.73970
#> 3  26.21970 26.21971 26.21971 26.21972 26.21972 26.21972 26.21972 26.21972
#> 4  19.99393 19.99402 19.99410 19.99416 19.99421 19.99425 19.99428 19.99430
#> 5  19.62617 19.62626 19.62633 19.62638 19.62642 19.62645 19.62648 19.62650
#> 6  13.55198 13.55200 13.55201 13.55202 13.55203 13.55204 13.55204 13.55205
#> 7  22.88813 22.88797 22.88784 22.88774 22.88766 22.88760 22.88755 22.88751
#> 8  22.30910 22.30891 22.30876 22.30864 22.30855 22.30848 22.30842 22.30837
#> 9  19.30687 19.30694 19.30700 19.30704 19.30708 19.30710 19.30713 19.30714
#> 10 19.30687 19.30694 19.30700 19.30704 19.30708 19.30710 19.30713 19.30714
#> 11 13.88955 13.88959 13.88962 13.88964 13.88966 13.88968 13.88969 13.88971
#> 12 11.14372 11.14365 11.14360 11.14356 11.14352 11.14349 11.14347 11.14345
#> 13 10.61346 10.61337 10.61330 10.61324 10.61320 10.61316 10.61313 10.61311
#> 14 26.98459 26.98463 26.98466 26.98469 26.98471 26.98472 26.98474 26.98474
#> 15 23.57518 23.57504 23.57493 23.57484 23.57477 23.57472 23.57468 23.57464
#> 16 15.69574 15.69585 15.69594 15.69601 15.69607 15.69612 15.69615 15.69618
#> 17 27.48975 27.48981 27.48986 27.48990 27.48993 27.48995 27.48997 27.48998
#> 18 25.41142 25.41121 25.41104 25.41091 25.41080 25.41072 25.41065 25.41060
#> 19 16.22589 16.22605 16.22618 16.22628 16.22636 16.22642 16.22647 16.22651
#> 20 20.64908 20.64904 20.64902 20.64900 20.64898 20.64897 20.64896 20.64895
#> 21 14.05353 14.05360 14.05365 14.05370 14.05373 14.05376 14.05378 14.05379
#>      0.0945   0.0881   0.0822   0.0767   0.0715   0.0667   0.0622   0.0580
#> 1  22.16786 21.84714 21.92859 21.92838 21.92820 21.92805 21.92791 21.92780
#> 2  21.68745 21.39985 21.61368 21.61353 21.61340 21.61329 21.61319 21.61310
#> 3  26.28782 26.30746 26.06959 26.06961 26.06963 26.06966 26.06969 26.06971
#> 4  20.03167 20.08127 20.01605 20.01609 20.01611 20.01613 20.01614 20.01615
#> 5  19.66470 19.74363 19.89928 19.89939 19.89948 19.89955 19.89961 19.89965
#> 6  13.57555 13.35086 13.13456 13.13459 13.13463 13.13466 13.13469 13.13472
#> 7  22.84866 22.97980 22.64864 22.64838 22.64817 22.64800 22.64786 22.64774
#> 8  22.29966 22.44209 23.36793 23.36835 23.36871 23.36902 23.36927 23.36948
#> 9  19.20722 19.16318 18.83548 18.83514 18.83485 18.83460 18.83439 18.83421
#> 10 19.20722 19.16318 19.08361 19.08338 19.08319 19.08301 19.08286 19.08273
#> 11 13.91489 13.95453 14.04910 14.04916 14.04921 14.04925 14.04928 14.04930
#> 12 11.16728 11.14084 11.10485 11.10481 11.10479 11.10477 11.10475 11.10474
#> 13 10.65027 10.65143 10.49434 10.49428 10.49424 10.49422 10.49420 10.49419
#> 14 27.02473 27.01530 27.14530 27.14539 27.14546 27.14552 27.14557 27.14561
#> 15 23.60387 23.32400 23.15734 23.15732 23.15732 23.15732 23.15734 23.15735
#> 16 15.73033 15.90435 16.10956 16.10969 16.10979 16.10987 16.10993 16.10998
#> 17 27.52398 27.48013 27.47751 27.47754 27.47756 27.47758 27.47759 27.47760
#> 18 25.44797 25.85969 25.58663 25.58660 25.58659 25.58658 25.58659 25.58660
#> 19 16.30475 16.92408 16.89959 16.89997 16.90029 16.90055 16.90077 16.90095
#> 20 20.55240 20.38610 20.31832 20.31812 20.31796 20.31782 20.31770 20.31760
#> 21 14.00172 13.78110 13.96003 13.96024 13.96041 13.96056 13.96068 13.96079
#>      0.0541   0.0504   0.0470   0.0439   0.0409   0.0382   0.0356   0.0332
#> 1  21.92770 21.92760 21.67403 21.67392 21.67380 21.67366 21.71115 21.71110
#> 2  21.61302 21.61297 21.60306 21.60302 21.60299 21.60296 21.66342 21.66342
#> 3  26.06974 26.06965 26.26232 26.26244 26.26256 26.26206 26.24116 26.24109
#> 4  20.01615 20.01601 19.49475 19.49463 19.49451 19.49367 19.38242 19.38250
#> 5  19.89968 19.89969 19.96844 19.96850 19.96856 19.96719 19.77070 19.77088
#> 6  13.13475 13.13493 12.89490 12.89474 12.89458 12.89485 12.93051 12.93032
#> 7  22.64765 22.64744 22.51591 22.51592 22.51592 22.51587 22.43873 22.43861
#> 8  23.36966 23.36993 23.41178 23.41167 23.41156 23.41188 23.45754 23.45766
#> 9  18.83406 18.83400 18.83475 18.83486 18.83496 18.83629 18.92961 18.92960
#> 10 19.08262 19.08262 19.04269 19.04276 19.04282 19.04406 19.14025 19.14030
#> 11 14.04932 14.04957 15.04495 15.04527 15.04557 15.04610 15.06978 15.06991
#> 12 11.10474 11.10452 10.72691 10.72678 10.72665 10.72614 10.72228 10.72219
#> 13 10.49419 10.49404 10.57604 10.57607 10.57611 10.57609 10.62399 10.62386
#> 14 27.14564 27.14559 27.24661 27.24667 27.24673 27.24669 27.30047 27.30052
#> 15 23.15738 23.15755 23.27292 23.27288 23.27284 23.27338 23.34843 23.34813
#> 16 16.11000 16.11005 16.11606 16.11604 16.11602 16.11607 16.08613 16.08637
#> 17 27.47761 27.47750 27.32065 27.32064 27.32062 27.32044 27.35027 27.35028
#> 18 25.58661 25.58644 25.74460 25.74463 25.74466 25.74496 25.74136 25.74128
#> 19 16.90110 16.90107 16.67570 16.67569 16.67568 16.67649 16.76497 16.76523
#> 20 20.31751 20.31758 20.43128 20.43130 20.43131 20.43024 20.25234 20.25229
#> 21 13.96087 13.96127 14.04162 14.04159 14.04155 14.04092 13.97447 13.97443
#>      0.0309   0.0289   0.0269   0.0251   0.0234   0.0218   0.0204   0.0190
#> 1  21.68763 21.68773 21.68782 21.68791 21.68800 21.68808 21.68816 21.68823
#> 2  21.61620 21.61621 21.61622 21.61624 21.61625 21.61626 21.61626 21.61627
#> 3  26.30485 26.30476 26.30468 26.30459 26.30451 26.30444 26.30436 26.30429
#> 4  19.47155 19.47171 19.47187 19.47201 19.47215 19.47229 19.47242 19.47255
#> 5  19.80504 19.80503 19.80502 19.80501 19.80501 19.80500 19.80500 19.80499
#> 6  12.91563 12.91577 12.91591 12.91604 12.91617 12.91629 12.91641 12.91653
#> 7  22.45749 22.45752 22.45754 22.45757 22.45759 22.45761 22.45763 22.45765
#> 8  23.32231 23.32236 23.32241 23.32246 23.32251 23.32255 23.32260 23.32264
#> 9  19.04620 19.04610 19.04601 19.04592 19.04583 19.04574 19.04566 19.04558
#> 10 19.23177 19.23171 19.23164 19.23157 19.23151 19.23145 19.23139 19.23133
#> 11 14.96937 14.96905 14.96875 14.96846 14.96818 14.96791 14.96765 14.96740
#> 12 10.72833 10.72846 10.72858 10.72870 10.72881 10.72892 10.72902 10.72912
#> 13 10.61948 10.61944 10.61940 10.61936 10.61932 10.61929 10.61925 10.61922
#> 14 27.33669 27.33662 27.33655 27.33649 27.33642 27.33636 27.33631 27.33625
#> 15 23.31480 23.31482 23.31485 23.31487 23.31489 23.31491 23.31493 23.31495
#> 16 16.01790 16.01791 16.01792 16.01792 16.01793 16.01794 16.01794 16.01795
#> 17 27.41084 27.41086 27.41088 27.41090 27.41091 27.41093 27.41094 27.41096
#> 18 25.61638 25.61634 25.61630 25.61626 25.61623 25.61619 25.61616 25.61613
#> 19 16.81483 16.81482 16.81481 16.81480 16.81479 16.81478 16.81477 16.81476
#> 20 20.21588 20.21591 20.21594 20.21597 20.21599 20.21602 20.21604 20.21607
#> 21 13.99682 13.99687 13.99691 13.99696 13.99700 13.99705 13.99709 13.99713
#>      0.0177   0.0165   0.0154   0.0144   0.0134   0.0125   0.0117   0.0109
#> 1  21.68830 21.68837 21.68844 21.68850 21.68856 21.68862 21.68868 21.68873
#> 2  21.61628 21.61629 21.61629 21.61630 21.61631 21.61631 21.61632 21.61633
#> 3  26.30423 26.30416 26.30410 26.30404 26.30398 26.30392 26.30387 26.30382
#> 4  19.47267 19.47279 19.47290 19.47301 19.47311 19.47321 19.47331 19.47340
#> 5  19.80499 19.80498 19.80498 19.80497 19.80497 19.80497 19.80496 19.80496
#> 6  12.91664 12.91675 12.91685 12.91695 12.91704 12.91714 12.91722 12.91731
#> 7  22.45767 22.45769 22.45771 22.45773 22.45774 22.45776 22.45777 22.45779
#> 8  23.32268 23.32272 23.32276 23.32280 23.32284 23.32287 23.32291 23.32294
#> 9  19.04550 19.04542 19.04535 19.04528 19.04521 19.04515 19.04508 19.04502
#> 10 19.23128 19.23122 19.23117 19.23112 19.23107 19.23103 19.23098 19.23094
#> 11 14.96716 14.96693 14.96671 14.96649 14.96629 14.96609 14.96590 14.96571
#> 12 10.72922 10.72931 10.72940 10.72948 10.72957 10.72965 10.72972 10.72980
#> 13 10.61919 10.61915 10.61912 10.61910 10.61907 10.61904 10.61902 10.61899
#> 14 27.33620 27.33614 27.33609 27.33605 27.33600 27.33596 27.33591 27.33587
#> 15 23.31497 23.31498 23.31500 23.31502 23.31503 23.31505 23.31507 23.31508
#> 16 16.01796 16.01796 16.01797 16.01797 16.01798 16.01799 16.01799 16.01800
#> 17 27.41097 27.41099 27.41100 27.41101 27.41102 27.41103 27.41105 27.41106
#> 18 25.61610 25.61607 25.61604 25.61601 25.61598 25.61596 25.61594 25.61591
#> 19 16.81476 16.81475 16.81474 16.81474 16.81473 16.81473 16.81472 16.81472
#> 20 20.21609 20.21611 20.21613 20.21615 20.21617 20.21619 20.21621 20.21622
#> 21 13.99717 13.99721 13.99724 13.99728 13.99731 13.99735 13.99738 13.99741
#>      0.0101   0.0095   0.0088   0.0082   0.0077   0.0071   0.0067   0.0062
#> 1  21.68878 21.68883 21.68888 21.68893 21.68897 21.68901 21.68906 21.68910
#> 2  21.61633 21.61634 21.61634 21.61634 21.61635 21.61635 21.61636 21.61636
#> 3  26.30377 26.30372 26.30368 26.30363 26.30359 26.30355 26.30351 26.30347
#> 4  19.47349 19.47358 19.47366 19.47374 19.47381 19.47389 19.47396 19.47402
#> 5  19.80496 19.80495 19.80495 19.80495 19.80495 19.80494 19.80494 19.80494
#> 6  12.91739 12.91747 12.91755 12.91762 12.91769 12.91776 12.91782 12.91789
#> 7  22.45780 22.45781 22.45783 22.45784 22.45785 22.45786 22.45787 22.45788
#> 8  23.32297 23.32300 23.32303 23.32306 23.32309 23.32312 23.32314 23.32317
#> 9  19.04496 19.04490 19.04485 19.04480 19.04475 19.04470 19.04465 19.04461
#> 10 19.23090 19.23086 19.23082 19.23078 19.23074 19.23071 19.23068 19.23064
#> 11 14.96554 14.96537 14.96521 14.96505 14.96490 14.96475 14.96461 14.96448
#> 12 10.72987 10.72993 10.73000 10.73006 10.73012 10.73018 10.73024 10.73029
#> 13 10.61897 10.61894 10.61892 10.61890 10.61888 10.61886 10.61884 10.61882
#> 14 27.33583 27.33579 27.33576 27.33572 27.33569 27.33566 27.33563 27.33560
#> 15 23.31509 23.31511 23.31512 23.31514 23.31515 23.31516 23.31517 23.31518
#> 16 16.01800 16.01800 16.01801 16.01801 16.01802 16.01802 16.01802 16.01803
#> 17 27.41107 27.41108 27.41109 27.41109 27.41110 27.41111 27.41112 27.41113
#> 18 25.61589 25.61587 25.61585 25.61583 25.61581 25.61579 25.61577 25.61576
#> 19 16.81471 16.81470 16.81470 16.81469 16.81469 16.81469 16.81468 16.81468
#> 20 20.21624 20.21626 20.21627 20.21629 20.21630 20.21631 20.21633 20.21634
#> 21 13.99744 13.99747 13.99750 13.99752 13.99755 13.99758 13.99760 13.99762
#>      0.0058   0.0054   0.0050   0.0047
#> 1  21.68913 21.68917 21.68920 21.68924
#> 2  21.61636 21.61637 21.61637 21.61637
#> 3  26.30344 26.30340 26.30337 26.30334
#> 4  19.47409 19.47415 19.47421 19.47427
#> 5  19.80493 19.80493 19.80493 19.80493
#> 6  12.91795 12.91800 12.91806 12.91811
#> 7  22.45789 22.45790 22.45791 22.45792
#> 8  23.32319 23.32321 23.32323 23.32325
#> 9  19.04456 19.04452 19.04448 19.04444
#> 10 19.23061 19.23058 19.23055 19.23053
#> 11 14.96435 14.96423 14.96411 14.96399
#> 12 10.73034 10.73039 10.73044 10.73049
#> 13 10.61881 10.61879 10.61878 10.61876
#> 14 27.33557 27.33554 27.33551 27.33549
#> 15 23.31520 23.31521 23.31522 23.31523
#> 16 16.01803 16.01803 16.01803 16.01804
#> 17 27.41113 27.41114 27.41115 27.41115
#> 18 25.61574 25.61572 25.61571 25.61570
#> 19 16.81467 16.81467 16.81466 16.81466
#> 20 20.21635 20.21636 20.21638 20.21639
#> 21 13.99765 13.99767 13.99769 13.99771
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 22.8 21.4 18.1 14.3 24.4 22.8 19.2 17.8 16.4 10.4 10.4 32.4 21.5
#> [16] 15.2 27.3 26.0 15.8 19.7 15.0
#> 
#> $X
#>               am       carb       cyl        disp         drat       gear
#>  [1,]  1.1547005  0.5477226  0.000000 -0.51315701  0.608676244  0.2603778
#>  [2,]  1.1547005  0.5477226  0.000000 -0.51315701  0.608676244  0.2603778
#>  [3,]  1.1547005 -1.0954451 -1.224745 -0.96484769  0.500900518  0.2603778
#>  [4,] -0.8660254 -1.0954451  0.000000  0.33810619 -1.158845666 -1.1066057
#>  [5,] -0.8660254 -1.0954451  0.000000  0.05145634 -1.848610314 -1.1066057
#>  [6,] -0.8660254  0.5477226  1.224745  1.22411483 -0.878628778 -1.1066057
#>  [7,] -0.8660254 -0.5477226 -1.224745 -0.62868559  0.156018194  0.2603778
#>  [8,] -0.8660254 -0.5477226 -1.224745 -0.67993511  0.651786535  0.2603778
#>  [9,] -0.8660254  0.5477226  0.000000 -0.44714068  0.651786535  0.2603778
#> [10,] -0.8660254  0.5477226  0.000000 -0.44714068  0.651786535  0.2603778
#> [11,] -0.8660254  0.0000000  1.224745  0.49272338 -1.180400811 -1.1066057
#> [12,] -0.8660254  0.5477226  1.224745  2.19698706 -1.482172844 -1.1066057
#> [13,] -0.8660254  0.5477226  1.224745  2.09275075 -1.331286828 -1.1066057
#> [14,]  1.1547005 -1.0954451 -1.224745 -1.21935801  0.996668859  0.2603778
#> [15,] -0.8660254 -1.0954451 -1.224745 -0.85974274  0.177573339 -1.1066057
#> [16,] -0.8660254 -0.5477226  1.224745  0.73767871 -1.007959649 -1.1066057
#> [17,]  1.1547005 -1.0954451 -1.224745 -1.21675211  0.996668859  0.2603778
#> [18,]  1.1547005 -0.5477226 -1.224745 -0.85800547  1.751098942  1.6273614
#> [19,]  1.1547005  0.5477226  1.224745  1.14593760  1.298440892  1.6273614
#> [20,]  1.1547005  1.6431677  0.000000 -0.64345240  0.005132177  1.6273614
#> [21,]  1.1547005  2.7386128  1.224745  0.71161964 -0.167308985  1.6273614
#>                hp         qsec         vs          wt
#>  [1,] -0.47382806 -0.767248753 -0.9534626 -0.69064291
#>  [2,] -0.47382806 -0.487123160 -0.9534626 -0.39917843
#>  [3,] -0.71207681  0.308233432  1.0488088 -1.03354229
#>  [4,] -0.47382806  0.723419578  1.0488088 -0.01055912
#>  [5,] -0.54390122  1.113594511  1.0488088  0.26947537
#>  [6,]  1.41814736 -1.077387801 -0.9534626  0.39520515
#>  [7,] -1.14653043  1.003545171  1.0488088 -0.03913407
#>  [8,] -0.68404755  2.454195560  1.0488088 -0.08485399
#>  [9,] -0.29163783  0.153163908  1.0488088  0.24661542
#> [10,] -0.29163783  0.453298471  1.0488088  0.24661542
#> [11,]  0.50719623 -0.297037937 -0.9534626  0.96670413
#> [12,]  0.85756205 -0.006907859 -0.9534626  2.31544171
#> [13,]  0.99770837 -0.086943743 -0.9534626  2.51432335
#> [14,] -1.09047190  0.738426306  1.0488088 -1.17070205
#> [15,] -0.65601828  1.008547413  1.0488088 -0.86780759
#> [16,]  0.08675725 -0.347060364 -0.9534626  0.24090043
#> [17,] -1.09047190  0.453298471  1.0488088 -1.47359650
#> [18,] -0.74010608 -0.647194927 -0.9534626 -1.23928192
#> [19,]  1.68442538 -1.747688326 -0.9534626 -0.06199403
#> [20,]  0.43712307 -1.247464054 -0.9534626 -0.51919321
#> [21,]  2.67946430 -1.697665899 -0.9534626  0.39520515
#> attr(,"center")
#>  [1]   0.4285714   3.0000000   6.0000000 219.0761905   3.6176190   3.8095238
#>  [7] 143.8095238  17.9938095   0.4761905   3.2242381
#> attr(,"scale")
#>  [1]   0.4948717   1.8257419   1.6329932 115.1230311   0.4639264   0.7315377
#>  [7]  71.3539928   1.9991033   0.4994328   0.8748922
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 39
#> 
#> $lambda.min
#> [1] 0.3318193
#> 
#> $null.dev
#> [1] 27.92998
#> 
#> $Bias
#> [1] 2.794544
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 11.98243