Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 34.41374 31.73633 27.31573 23.77582 21.02179 18.92939 17.39203 16.31838
#>   [9] 15.63027 15.26094 15.15342 15.25919 15.53706 15.95208 16.47470 17.00823
#>  [17] 16.76790 16.58425 16.49523 16.41429 16.34061 16.27346 16.21221 16.15628
#>  [25] 16.10634 16.24984 16.08261 15.84311 14.98821 13.68554 13.69398 13.72233
#>  [33] 13.75399 13.79420 13.80954 13.81351 14.27700 14.22555 15.32229 15.26314
#>  [41] 15.20803 31.33574 31.31941 31.68979 32.56620 30.37175 30.35694 31.27424
#>  [49] 30.24057 33.82717 33.82295 33.06293 32.10786 30.68241 29.87836 36.93369
#>  [57] 36.58911 36.41933 36.40874 38.12088 38.12249 38.12409 38.12566 37.90867
#>  [65] 37.91011 35.08143 34.80250 34.80346 34.80440 34.80534 34.80627 34.80720
#>  [73] 34.80811 34.81684 34.81770 34.81855 34.81939 34.82022 34.82103 34.82183
#>  [81] 34.82262 34.82340 34.82416 34.82491 34.79675 34.79741 34.79806 34.79869
#>  [89] 34.79932 34.79993 34.80053 34.80112 34.80170 34.80226 34.80282 34.80337
#>  [97] 34.80440 34.41619 34.41558 34.41497
#> 
#> $cvse
#>   [1] 10.233211 10.128150  9.296407  8.506649  7.780441  7.117222  6.527060
#>   [8]  6.028016  5.642902  5.394279  5.297656  5.355279  5.554411  5.871394
#>  [15]  6.278452  6.691921  6.618752  6.553790  6.506832  6.463360  6.423083
#>  [22]  6.385737  6.351086  6.318918  6.288907  6.236562  6.214085  6.198696
#>  [29]  6.181015  5.890821  5.858872  5.834284  5.811027  5.800303  5.782220
#>  [36]  5.764586  5.735785  5.723505  5.743401  5.734308  5.726187 12.211913
#>  [43] 12.212485 12.178743 12.020121 11.735933 11.724087 11.653514 11.704786
#>  [50] 11.694437 11.654666 11.655910 11.738866 11.326636 11.371041 12.911500
#>  [57] 12.919935 12.940430 12.940917 13.075379 13.076080 13.076768 13.077443
#>  [64] 13.093512 13.094169 13.156280 13.165327 13.165930 13.166522 13.167104
#>  [71] 13.167674 13.168234 13.168784 13.170093 13.170623 13.171143 13.171653
#>  [78] 13.172154 13.172644 13.173126 13.173597 13.174060 13.174514 13.174958
#>  [85] 13.168514 13.168927 13.169331 13.169727 13.170115 13.170495 13.170867
#>  [92] 13.171232 13.171589 13.171938 13.172280 13.172616 13.173130 12.988418
#>  [99] 12.988206 12.987993
#> 
#> $fold
#>  [1]  9 10  2  3  9  6  1  5  4  3  1 10  7  4  8  6  8  2  5  1  7
#> 
#> $lambda
#>   [1] 4.889270570 4.559750098 4.252438202 3.965838100 3.698553885 3.449283732
#>   [7] 3.216813552 3.000011085 2.797820379 2.609256649 2.433401484 2.269398368
#>  [13] 2.116448514 1.973806967 1.840778984 1.716716641 1.601015685 1.493112586
#>  [19] 1.392481795 1.298633183 1.211109653 1.129484916 1.053361413 0.982368379
#>  [25] 0.916160038 0.854413918 0.796829279 0.743125653 0.693041471 0.646332795
#>  [31] 0.602772128 0.562147304 0.524260457 0.488927057 0.455975010 0.425243820
#>  [37] 0.396583810 0.369855388 0.344928373 0.321681355 0.300001109 0.279782038
#>  [43] 0.260925665 0.243340148 0.226939837 0.211644851 0.197380697 0.184077898
#>  [49] 0.171671664 0.160101569 0.149311259 0.139248180 0.129863318 0.121110965
#>  [55] 0.112948492 0.105336141 0.098236838 0.091616004 0.085441392 0.079682928
#>  [61] 0.074312565 0.069304147 0.064633280 0.060277213 0.056214730 0.052426046
#>  [67] 0.048892706 0.045597501 0.042524382 0.039658381 0.036985539 0.034492837
#>  [73] 0.032168136 0.030000111 0.027978204 0.026092566 0.024334015 0.022693984
#>  [79] 0.021164485 0.019738070 0.018407790 0.017167166 0.016010157 0.014931126
#>  [85] 0.013924818 0.012986332 0.012111097 0.011294849 0.010533614 0.009823684
#>  [91] 0.009161600 0.008544139 0.007968293 0.007431257 0.006930415 0.006463328
#>  [97] 0.006027721 0.005621473 0.005242605 0.004889271
#> 
#> $fit
#> $beta
#>              4.8893     4.5598     4.2524     3.9658    3.6986    3.4493
#> (Intercept) 21.2619 22.9913250 24.6041880 26.1083495 27.511136 28.819379
#> am           0.0000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> carb         0.0000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> cyl          0.0000 -0.2928857 -0.5660318 -0.8207689 -1.058337 -1.279895
#> disp         0.0000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> drat         0.0000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> gear         0.0000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> hp           0.0000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> qsec         0.0000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> vs           0.0000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> wt           0.0000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#>               3.2168    3.0000    2.7978    2.6093    2.4334    2.2694
#> (Intercept) 30.03945 31.177293 32.238449 33.228087 34.151027 35.011763
#> am           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.48652 -1.679219 -1.858931 -2.026531 -2.182835 -2.328605
#> disp         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                2.1164    1.9738    1.8408    1.7167    1.6010    1.4931
#> (Intercept) 35.814489 36.563114 37.261284 37.912399 38.368790 38.368790
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.464551 -2.591334 -2.709572 -2.819842 -2.897134 -2.897134
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.3925    1.2986    1.2111    1.1295    1.0534    0.9824
#> (Intercept) 38.368790 38.368790 38.368790 38.368790 38.368790 38.368790
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.897134 -2.897134 -2.897134 -2.897134 -2.897134 -2.897134
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                0.9162    0.8544    0.7968    0.7431    0.6930    0.6463
#> (Intercept) 38.368790 38.368790 40.013107 40.067370 40.067170 40.067038
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.897134 -2.897134 -1.788713 -1.751669 -1.751481 -1.751358
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000 -2.716049 -2.806591 -2.806893 -2.807091
#>                0.6028      0.5621      0.5243     0.4889     0.4560     0.4252
#> (Intercept) 40.066953 40.05780667 40.04443487 40.0319403 40.0204689 40.0096004
#> am           0.000000  0.00000000  0.00000000  0.0000000  0.0000000  0.0000000
#> carb         0.000000 -0.03481562 -0.08463739 -0.1311128 -0.1743243 -0.2147477
#> cyl         -1.751277 -1.73034328 -1.70011933 -1.6719058 -1.6458105 -1.6212700
#> disp         0.000000  0.00000000  0.00000000  0.0000000  0.0000000  0.0000000
#> drat         0.000000  0.00000000  0.00000000  0.0000000  0.0000000  0.0000000
#> gear         0.000000  0.00000000  0.00000000  0.0000000  0.0000000  0.0000000
#> hp           0.000000  0.00000000  0.00000000  0.0000000  0.0000000  0.0000000
#> qsec         0.000000  0.00000000  0.00000000  0.0000000  0.0000000  0.0000000
#> vs           0.000000  0.00000000  0.00000000  0.0000000  0.0000000  0.0000000
#> wt          -2.807221 -2.81384255 -2.82374687 -2.8330175 -2.8414172 -2.8494821
#>                 0.3966     0.3699     0.3449     0.3217     0.3000     0.2798
#> (Intercept) 39.9994531 39.9899891 39.9811628 39.9729314 23.8247577 23.8222234
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb        -0.2524543 -0.2876201 -0.3204159 -0.3510013 -0.8100507 -0.8101583
#> cyl         -1.5983702 -1.5770129 -1.5570949 -1.5385194 -0.6232960 -0.6230275
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  3.2093740  3.2098527
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.8570188 -2.8640486 -2.8706046 -2.8767188 -2.7360900 -2.7362504
#>                 0.2609     0.2433    0.2269     0.2116     0.1974     0.1841
#> (Intercept) 23.8201347 23.8184132 16.828516 16.8308823 16.4459692 16.4440323
#> am           0.0000000  0.0000000  0.000000  0.0000000  1.8658631  1.8666075
#> carb        -0.8102470 -0.8103202 -0.780532 -0.7806149 -0.9398432 -0.9399297
#> cyl         -0.6228063 -0.6226240  0.000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> drat         3.2102473  3.2105725  4.080318  4.0799783  3.3972479  3.3973162
#> gear         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> vs           0.0000000  0.0000000  1.622510  1.6219745  1.8381883  1.8383971
#> wt          -2.7363826 -2.7364916 -2.960170 -2.9603901 -2.2316622 -2.2311858
#>                 0.1717       0.1601       0.1493       0.1392       0.1299
#> (Intercept) 16.4425857 -12.47045031 -12.47365332 -12.47635170 -12.47860860
#> am           1.8671633   4.10212640   4.10306571   4.10387678   4.10457421
#> carb        -0.9399943   0.00000000   0.00000000   0.00000000   0.00000000
#> cyl          0.0000000   0.00000000   0.00000000   0.00000000   0.00000000
#> disp         0.0000000   0.05640720   0.05641014   0.05641240   0.05641408
#> drat         3.3973672   3.03694985   3.03660499   3.03630736   3.03605160
#> gear         0.0000000   2.12123355   2.12117411   2.12110461   2.12102764
#> hp           0.0000000  -0.03670337  -0.03670053  -0.03669784  -0.03669531
#> qsec         0.0000000   1.12756306   1.12778414   1.12797550   1.12814049
#> vs           1.8385532   2.52759459   2.52739949   2.52722035   2.52705620
#> wt          -2.2308301  -5.10721449  -5.10744087  -5.10762860  -5.10778276
#>                   0.1211       0.1129       0.1053       0.0982       0.0916
#> (Intercept) -12.48048042 -12.48201744 -40.72503902 -40.71998140 -40.71496544
#> am            4.10517117   4.10567952   5.56038797   5.56018383   5.55996810
#> carb          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> cyl           0.00000000   0.00000000   1.83295568   1.83263104   1.83231491
#> disp          0.05641527   0.05641604   0.06458880   0.06458676   0.06458458
#> drat          3.03583288   3.03564683   4.25675259   4.25650051   4.25626060
#> gear          2.12094543   2.12085983   3.53948425   3.53918406   3.53888445
#> hp           -0.03669293  -0.03669071  -0.05425787  -0.05425405  -0.05425027
#> qsec          1.12828214   1.12840319   1.59254837   1.59248059   1.59241089
#> vs            2.52690607   2.52676903   4.21715341   4.21680562   4.21646498
#> wt           -5.10790788  -5.10800794  -5.61356726  -5.61347011  -5.61336791
#>                   0.0854       0.0797       0.0743       0.0693       0.0646
#> (Intercept) -40.70999961 -40.70509158 -40.70024831 -40.69547606 -40.69078042
#> am            5.55974220   5.55950746   5.55926511   5.55901631   5.55876216
#> carb          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> cyl           1.83200736   1.83170844   1.83141817   1.83113656   1.83086358
#> disp          0.06458229   0.06457990   0.06457742   0.06457487   0.06457225
#> drat          4.25603246   4.25581571   4.25560996   4.25541484   4.25522996
#> gear          3.53858607   3.53828953   3.53799537   3.53770410   3.53741617
#> hp           -0.05424655  -0.05424288  -0.05423927  -0.05423572  -0.05423224
#> qsec          1.59233959   1.59226698   1.59219333   1.59211889   1.59204391
#> vs            4.21613172   4.21580604   4.21548812   4.21517808   4.21487604
#> wt           -5.61326129  -5.61315086  -5.61303717  -5.61292073  -5.61280205
#>                   0.0603       0.0562       0.0524       0.0489       0.0456
#> (Intercept) -40.68616639 -40.68163835 -40.67720017 -40.67285519 -40.66860627
#> am            5.55850366   5.55824177   5.55797735   5.55771123   5.55744417
#> carb          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> cyl           1.83059920   1.83034336   1.83009599   1.82985701   1.82962632
#> disp          0.06456958   0.06456688   0.06456414   0.06456137   0.06455860
#> drat          4.25505495   4.25488946   4.25473312   4.25458558   4.25444650
#> gear          3.53713200   3.53685195   3.53657635   3.53630550   3.53603967
#> hp           -0.05422883  -0.05422549  -0.05422222  -0.05421903  -0.05421592
#> qsec          1.59196860   1.59189316   1.59181777   1.59174262   1.59166785
#> vs            4.21458207   4.21429623   4.21401856   4.21374905   4.21348771
#> wt           -5.61268156  -5.61255970  -5.61243686  -5.61231340  -5.61218965
#>                   0.0425       0.0397       0.0370       0.0345       0.0322
#> (Intercept) -40.66445582 -40.66040585 -44.12418429 -44.12953609 -44.13473083
#> am            5.55717685   5.55690993   5.56195979   5.56226369   5.56256100
#> carb          0.00000000   0.00000000  -0.39945148  -0.39955967  -0.39966281
#> cyl           1.82940381   1.82918935   2.14242441   2.14275260   2.14306993
#> disp          0.06455581   0.06455303   0.05807760   0.05807856   0.05807954
#> drat          4.25431554   4.25419238   4.46504872   4.46521915   4.46538315
#> gear          3.53577908   3.53552393   3.95675575   3.95716089   3.95755335
#> hp           -0.05421289  -0.05420993  -0.04817264  -0.04817444  -0.04817623
#> qsec          1.59159362   1.59152005   1.54973016   1.54979553   1.54985973
#> vs            4.21323451   4.21298939   4.16413319   4.16439786   4.16465554
#> wt           -5.61206594  -5.61194254  -5.07969199  -5.07964366  -5.07960046
#>                   0.0300       0.0280       0.0261       0.0243       0.0227
#> (Intercept) -44.13977204 -44.14466321 -44.14940776 -44.15400907 -44.15847049
#> am            5.56285180   5.56313616   5.56341417   5.56368591   5.56395146
#> carb         -0.39976106  -0.39985460  -0.39994358  -0.40002817  -0.40010852
#> cyl           2.14337670   2.14367318   2.14395964   2.14423635   2.14450356
#> disp          0.05808056   0.05808159   0.05808265   0.05808373   0.05808482
#> drat          4.46554090   4.46569259   4.46583841   4.46597851   4.46611309
#> gear          3.95793342   3.95830142   3.95865764   3.95900238   3.95933592
#> hp           -0.04817800  -0.04817975  -0.04818149  -0.04818321  -0.04818491
#> qsec          1.54992276   1.54998463   1.55004533   1.55010489   1.55016329
#> vs            4.16490635   4.16515043   4.16538792   4.16561893   4.16584361
#> wt           -5.07956215  -5.07952851  -5.07949932  -5.07947438  -5.07945349
#>                   0.0212       0.0197      0.0184       0.0172       0.0160
#> (Intercept) -44.16279531 -44.16698677 -44.1710481 -44.17498232 -44.17879265
#> am            5.56421089   5.56446431   5.5647118   5.56495340   5.56518926
#> carb         -0.40018477  -0.40025708  -0.4003256  -0.40039043  -0.40045174
#> cyl           2.14476153   2.14501052   2.1452508   2.14548251   2.14570599
#> disp          0.05808594   0.05808706   0.0580882   0.05808935   0.05809051
#> drat          4.46624231   4.46636632   4.4664853   4.46659940   4.46670877
#> gear          3.95965855   3.95997056   3.9602722   3.96056378   3.96084552
#> hp           -0.04818660  -0.04818826  -0.0481899  -0.04819153  -0.04819313
#> qsec          1.55022056   1.55027669   1.5503317   1.55038559   1.55043837
#> vs            4.16606207   4.16627446   4.1664809   4.16668148   4.16687637
#> wt           -5.07943645  -5.07942308  -5.0794132  -5.07940661  -5.07940318
#>                   0.0149       0.0139       0.0130       0.0121       0.0113
#> (Intercept) -44.18248210 -44.18605365 -44.18951026 -44.19285482 -44.19609016
#> am            5.56541944   5.56564404   5.56586314   5.56607682   5.56628519
#> carb         -0.40050964  -0.40056427  -0.40061575  -0.40066419  -0.40070971
#> cyl           2.14592144   2.14612908   2.14632914   2.14652182   2.14670735
#> disp          0.05809168   0.05809285   0.05809403   0.05809521   0.05809640
#> drat          4.46681356   4.46691391   4.46700998   4.46710189   4.46718978
#> gear          3.96111771   3.96138059   3.96163442   3.96187943   3.96211587
#> hp           -0.04819471  -0.04819628  -0.04819782  -0.04819933  -0.04820083
#> qsec          1.55049006   1.55054066   1.55059019   1.55063866   1.55068607
#> vs            4.16706567   4.16724951   4.16742801   4.16760128   4.16776944
#> wt           -5.07940272  -5.07940509  -5.07941012  -5.07941769  -5.07942764
#>                   0.0105       0.0098       0.0092       0.0085       0.0080
#> (Intercept) -44.19921910 -44.20224436 -44.20516866 -44.20799462 -44.21072486
#> am            5.56648833   5.56668632   5.56687927   5.56706726   5.56725037
#> carb         -0.40075242  -0.40079244  -0.40082987  -0.40086480  -0.40089735
#> cyl           2.14688593   2.14705776   2.14722304   2.14738197   2.14753474
#> disp          0.05809758   0.05809877   0.05809995   0.05810113   0.05810231
#> drat          4.46727378   4.46735402   4.46743063   4.46750372   4.46757342
#> gear          3.96234398   3.96256399   3.96277611   3.96298058   3.96317760
#> hp           -0.04820230  -0.04820375  -0.04820518  -0.04820658  -0.04820797
#> qsec          1.55073245   1.55077780   1.55082213   1.55086546   1.55090780
#> vs            4.16793260   4.16809089   4.16824441   4.16839327   4.16853759
#> wt           -5.07943984  -5.07945417  -5.07947049  -5.07948869  -5.07950865
#>                   0.0074       0.0069       0.0065       0.0060       0.0056
#> (Intercept) -44.21336190 -44.21590826 -44.21836638 -44.22073867 -44.22302746
#> am            5.56742870   5.56760234   5.56777138   5.56793590   5.56809599
#> carb         -0.40092760  -0.40095564  -0.40098158  -0.40100549  -0.40102747
#> cyl           2.14768152   2.14782250   2.14795786   2.14808777   2.14821240
#> disp          0.05810348   0.05810465   0.05810582   0.05810697   0.05810812
#> drat          4.46763983   4.46770308   4.46776327   4.46782051   4.46787489
#> gear          3.96336740   3.96355017   3.96372612   3.96389546   3.96405836
#> hp           -0.04820933  -0.04821066  -0.04821198  -0.04821327  -0.04821453
#> qsec          1.55094917   1.55098958   1.55102904   1.55106756   1.55110516
#> vs            4.16867746   4.16881300   4.16894432   4.16907150   4.16919466
#> wt           -5.07953026  -5.07955342  -5.07957802  -5.07960396  -5.07963115
#>                   0.0052      0.0049
#> (Intercept) -44.22523508 -44.2273638
#> am            5.56825173   5.5684032
#> carb         -0.40104758  -0.4010659
#> cyl           2.14833191   2.1484465
#> disp          0.05810926   0.0581104
#> drat          4.46792653   4.4679755
#> gear          3.96421503   3.9643656
#> hp           -0.04821578  -0.0482170
#> qsec          1.55114186   1.5511777
#> vs            4.16931389   4.1694293
#> wt           -5.07965950  -5.0796889
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1   1   1   1 250  11   1   1   1   8   8   8   7   7
#>  [37]   7   7   7   7  46   1   1   1  33   1  34   1   1 203   1   1   1   1
#>  [55]   1 141   1   1   1   1   1   1   1   1   1   1   1   1   1   1 156   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.889270570 4.559750098 4.252438202 3.965838100 3.698553885 3.449283732
#>   [7] 3.216813552 3.000011085 2.797820379 2.609256649 2.433401484 2.269398368
#>  [13] 2.116448514 1.973806967 1.840778984 1.716716641 1.601015685 1.493112586
#>  [19] 1.392481795 1.298633183 1.211109653 1.129484916 1.053361413 0.982368379
#>  [25] 0.916160038 0.854413918 0.796829279 0.743125653 0.693041471 0.646332795
#>  [31] 0.602772128 0.562147304 0.524260457 0.488927057 0.455975010 0.425243820
#>  [37] 0.396583810 0.369855388 0.344928373 0.321681355 0.300001109 0.279782038
#>  [43] 0.260925665 0.243340148 0.226939837 0.211644851 0.197380697 0.184077898
#>  [49] 0.171671664 0.160101569 0.149311259 0.139248180 0.129863318 0.121110965
#>  [55] 0.112948492 0.105336141 0.098236838 0.091616004 0.085441392 0.079682928
#>  [61] 0.074312565 0.069304147 0.064633280 0.060277213 0.056214730 0.052426046
#>  [67] 0.048892706 0.045597501 0.042524382 0.039658381 0.036985539 0.034492837
#>  [73] 0.032168136 0.030000111 0.027978204 0.026092566 0.024334015 0.022693984
#>  [79] 0.021164485 0.019738070 0.018407790 0.017167166 0.016010157 0.014931126
#>  [85] 0.013924818 0.012986332 0.012111097 0.011294849 0.010533614 0.009823684
#>  [91] 0.009161600 0.008544139 0.007968293 0.007431257 0.006930415 0.006463328
#>  [97] 0.006027721 0.005621473 0.005242605 0.004889271
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 26
#> 
#> $loss
#>   [1] 689.46952 593.09988 512.47234 445.32149 389.69150 343.89476 306.47611
#>   [8] 276.18166 251.93182 232.79781 217.98138 206.79715 198.65727 193.05823
#>  [15] 189.56928 187.82253 187.46522 187.46522 187.46522 187.46522 187.46522
#>  [22] 187.46522 187.46522 187.46522 187.46522 187.46522 146.20799 146.16420
#>  [29] 146.16419 146.16419 146.16419 144.67397 142.75887 141.20318 139.95657
#>  [36] 138.96472 138.19144 137.60241 137.16800 136.86283 124.66398 124.66397
#>  [43] 124.66397 124.66397 120.36676 120.36675 116.34614 116.34613 116.34613
#>  [50]  84.72771  84.72770  84.72769  84.72769  84.72768  84.72768  77.54447
#>  [57]  77.54446  77.54445  77.54444  77.54442  77.54441  77.54440  77.54439
#>  [64]  77.54438  77.54438  77.54437  77.54436  77.54435  77.54434  77.54433
#>  [71]  76.94596  76.94595  76.94594  76.94593  76.94592  76.94591  76.94590
#>  [78]  76.94590  76.94589  76.94588  76.94588  76.94587  76.94587  76.94586
#>  [85]  76.94586  76.94585  76.94585  76.94584  76.94584  76.94584  76.94583
#>  [92]  76.94583  76.94583  76.94582  76.94582  76.94582  76.94582  76.94582
#>  [99]  76.94581  76.94581
#> 
#> $linear.predictors
#>     4.8893   4.5598   4.2524   3.9658   3.6986   3.4493   3.2168   3.0000
#> 1  21.2619 21.23401 21.20800 21.18374 21.16111 21.14001 21.12033 21.10198
#> 2  21.2619 21.23401 21.20800 21.18374 21.16111 21.14001 21.12033 21.10198
#> 3  21.2619 21.81978 22.34006 22.82527 23.27779 23.69980 24.09337 24.46042
#> 4  21.2619 21.23401 21.20800 21.18374 21.16111 21.14001 21.12033 21.10198
#> 5  21.2619 21.23401 21.20800 21.18374 21.16111 21.14001 21.12033 21.10198
#> 6  21.2619 20.64824 20.07593 19.54220 19.04444 18.58022 18.14729 17.74354
#> 7  21.2619 21.81978 22.34006 22.82527 23.27779 23.69980 24.09337 24.46042
#> 8  21.2619 21.23401 21.20800 21.18374 21.16111 21.14001 21.12033 21.10198
#> 9  21.2619 20.64824 20.07593 19.54220 19.04444 18.58022 18.14729 17.74354
#> 10 21.2619 20.64824 20.07593 19.54220 19.04444 18.58022 18.14729 17.74354
#> 11 21.2619 21.81978 22.34006 22.82527 23.27779 23.69980 24.09337 24.46042
#> 12 21.2619 21.81978 22.34006 22.82527 23.27779 23.69980 24.09337 24.46042
#> 13 21.2619 20.64824 20.07593 19.54220 19.04444 18.58022 18.14729 17.74354
#> 14 21.2619 20.64824 20.07593 19.54220 19.04444 18.58022 18.14729 17.74354
#> 15 21.2619 20.64824 20.07593 19.54220 19.04444 18.58022 18.14729 17.74354
#> 16 21.2619 21.81978 22.34006 22.82527 23.27779 23.69980 24.09337 24.46042
#> 17 21.2619 21.81978 22.34006 22.82527 23.27779 23.69980 24.09337 24.46042
#> 18 21.2619 21.81978 22.34006 22.82527 23.27779 23.69980 24.09337 24.46042
#> 19 21.2619 21.23401 21.20800 21.18374 21.16111 21.14001 21.12033 21.10198
#> 20 21.2619 20.64824 20.07593 19.54220 19.04444 18.58022 18.14729 17.74354
#> 21 21.2619 21.81978 22.34006 22.82527 23.27779 23.69980 24.09337 24.46042
#>      2.7978   2.6093   2.4334   2.2694   2.1164   1.9738   1.8408   1.7167
#> 1  21.08486 21.06890 21.05402 21.04013 21.02719 21.01511 21.00385 20.99335
#> 2  21.08486 21.06890 21.05402 21.04013 21.02719 21.01511 21.00385 20.99335
#> 3  24.80273 25.12196 25.41969 25.69734 25.95629 26.19778 26.42299 26.63303
#> 4  21.08486 21.06890 21.05402 21.04013 21.02719 21.01511 21.00385 20.99335
#> 5  21.08486 21.06890 21.05402 21.04013 21.02719 21.01511 21.00385 20.99335
#> 6  17.36700 17.01584 16.68835 16.38292 16.09808 15.83244 15.58471 15.35366
#> 7  24.80273 25.12196 25.41969 25.69734 25.95629 26.19778 26.42299 26.63303
#> 8  21.08486 21.06890 21.05402 21.04013 21.02719 21.01511 21.00385 20.99335
#> 9  17.36700 17.01584 16.68835 16.38292 16.09808 15.83244 15.58471 15.35366
#> 10 17.36700 17.01584 16.68835 16.38292 16.09808 15.83244 15.58471 15.35366
#> 11 24.80273 25.12196 25.41969 25.69734 25.95629 26.19778 26.42299 26.63303
#> 12 24.80273 25.12196 25.41969 25.69734 25.95629 26.19778 26.42299 26.63303
#> 13 17.36700 17.01584 16.68835 16.38292 16.09808 15.83244 15.58471 15.35366
#> 14 17.36700 17.01584 16.68835 16.38292 16.09808 15.83244 15.58471 15.35366
#> 15 17.36700 17.01584 16.68835 16.38292 16.09808 15.83244 15.58471 15.35366
#> 16 24.80273 25.12196 25.41969 25.69734 25.95629 26.19778 26.42299 26.63303
#> 17 24.80273 25.12196 25.41969 25.69734 25.95629 26.19778 26.42299 26.63303
#> 18 24.80273 25.12196 25.41969 25.69734 25.95629 26.19778 26.42299 26.63303
#> 19 21.08486 21.06890 21.05402 21.04013 21.02719 21.01511 21.00385 20.99335
#> 20 17.36700 17.01584 16.68835 16.38292 16.09808 15.83244 15.58471 15.35366
#> 21 24.80273 25.12196 25.41969 25.69734 25.95629 26.19778 26.42299 26.63303
#>      1.6010   1.4931   1.3925   1.2986   1.2111   1.1295   1.0534   0.9824
#> 1  20.98599 20.98599 20.98599 20.98599 20.98599 20.98599 20.98599 20.98599
#> 2  20.98599 20.98599 20.98599 20.98599 20.98599 20.98599 20.98599 20.98599
#> 3  26.78025 26.78025 26.78025 26.78025 26.78025 26.78025 26.78025 26.78025
#> 4  20.98599 20.98599 20.98599 20.98599 20.98599 20.98599 20.98599 20.98599
#> 5  20.98599 20.98599 20.98599 20.98599 20.98599 20.98599 20.98599 20.98599
#> 6  15.19172 15.19172 15.19172 15.19172 15.19172 15.19172 15.19172 15.19172
#> 7  26.78025 26.78025 26.78025 26.78025 26.78025 26.78025 26.78025 26.78025
#> 8  20.98599 20.98599 20.98599 20.98599 20.98599 20.98599 20.98599 20.98599
#> 9  15.19172 15.19172 15.19172 15.19172 15.19172 15.19172 15.19172 15.19172
#> 10 15.19172 15.19172 15.19172 15.19172 15.19172 15.19172 15.19172 15.19172
#> 11 26.78025 26.78025 26.78025 26.78025 26.78025 26.78025 26.78025 26.78025
#> 12 26.78025 26.78025 26.78025 26.78025 26.78025 26.78025 26.78025 26.78025
#> 13 15.19172 15.19172 15.19172 15.19172 15.19172 15.19172 15.19172 15.19172
#> 14 15.19172 15.19172 15.19172 15.19172 15.19172 15.19172 15.19172 15.19172
#> 15 15.19172 15.19172 15.19172 15.19172 15.19172 15.19172 15.19172 15.19172
#> 16 26.78025 26.78025 26.78025 26.78025 26.78025 26.78025 26.78025 26.78025
#> 17 26.78025 26.78025 26.78025 26.78025 26.78025 26.78025 26.78025 26.78025
#> 18 26.78025 26.78025 26.78025 26.78025 26.78025 26.78025 26.78025 26.78025
#> 19 20.98599 20.98599 20.98599 20.98599 20.98599 20.98599 20.98599 20.98599
#> 20 15.19172 15.19172 15.19172 15.19172 15.19172 15.19172 15.19172 15.19172
#> 21 26.78025 26.78025 26.78025 26.78025 26.78025 26.78025 26.78025 26.78025
#>      0.9162   0.8544   0.7968   0.7431   0.6930   0.6463   0.6028   0.5621
#> 1  20.98599 20.98599 22.16478 22.20408 22.20422 22.20431 22.20437 22.16422
#> 2  20.98599 20.98599 21.47219 21.48840 21.48846 21.48850 21.48853 21.44669
#> 3  26.78025 26.78025 26.55702 26.54940 26.54925 26.54915 26.54909 26.57350
#> 4  20.98599 20.98599 20.54873 20.53416 20.53412 20.53409 20.53407 20.59443
#> 5  20.98599 20.98599 19.88330 19.84655 19.84643 19.84635 19.84630 19.90504
#> 6  15.19172 15.19172 16.00711 16.03448 16.03471 16.03486 16.03496 16.03038
#> 7  26.78025 26.78025 24.30270 24.21993 24.21953 24.21927 24.21910 24.20320
#> 8  20.98599 20.98599 19.93762 19.90268 19.90257 19.90250 19.90245 19.85687
#> 9  15.19172 15.19172 15.43673 15.44510 15.44526 15.44537 15.44544 15.47429
#> 10 15.19172 15.19172 11.18612 11.05278 11.05247 11.05227 11.05214 11.03581
#> 11 26.78025 26.78025 26.88295 26.88619 26.88608 26.88601 26.88596 26.91116
#> 12 26.78025 26.78025 27.87430 27.91060 27.91059 27.91059 27.91059 27.93822
#> 13 15.19172 15.19172 16.14291 16.17481 16.17505 16.17521 16.17532 16.24070
#> 14 15.19172 15.19172 16.37377 16.41337 16.41364 16.41382 16.41393 16.47988
#> 15 15.19172 15.19172 15.26019 15.26267 15.26281 15.26291 15.26297 15.32620
#> 16 26.78025 26.78025 27.60270 27.62994 27.62991 27.62988 27.62987 27.65683
#> 17 26.78025 26.78025 27.04591 27.05459 27.05449 27.05443 27.05439 27.04518
#> 18 26.78025 26.78025 28.74887 28.81432 28.81441 28.81448 28.81452 28.80946
#> 19 20.98599 20.98599 21.75737 21.78309 21.78319 21.78325 21.78329 21.67251
#> 20 15.19172 15.19172 16.00711 16.03448 16.03471 16.03486 16.03496 15.89112
#> 21 26.78025 26.78025 25.30764 25.25837 25.25808 25.25789 25.25777 25.24432
#>      0.5243   0.4889   0.4560   0.4252   0.3966   0.3699   0.3449   0.3217
#> 1  22.10695 22.05355 22.00380 21.95735 21.91403 21.87362 21.83595 21.80081
#> 2  21.38690 21.33113 21.27923 21.23073 21.18549 21.14329 21.10394 21.06724
#> 3  26.60823 26.64060 26.67081 26.69897 26.72523 26.74972 26.77256 26.79386
#> 4  20.68074 20.76124 20.83613 20.90615 20.97146 21.03238 21.08918 21.14216
#> 5  19.98892 20.06715 20.13998 20.20802 20.27149 20.33068 20.38589 20.43737
#> 6  16.02415 16.01837 16.01283 16.00780 16.00312 15.99875 15.99468 15.99088
#> 7  24.17988 24.15809 24.13811 24.11916 24.10145 24.08494 24.06955 24.05519
#> 8  19.79148 19.73047 19.67383 19.62077 19.57127 19.52510 19.48205 19.44190
#> 9  15.51580 15.55455 15.59045 15.62415 15.65560 15.68492 15.71227 15.73778
#> 10 11.01200 10.98976 10.96931 10.94997 10.93191 10.91507 10.89936 10.88471
#> 11 26.94708 26.98057 27.01178 27.04091 27.06808 27.09341 27.11704 27.13907
#> 12 27.97774 28.01462 28.04890 28.08097 28.11089 28.13879 28.16481 28.18907
#> 13 16.33462 16.42225 16.50355 16.57977 16.65088 16.71719 16.77904 16.83672
#> 14 16.57463 16.66305 16.74507 16.82197 16.89372 16.96064 17.02304 17.08124
#> 15 15.41690 15.50152 15.58009 15.65369 15.72235 15.78638 15.84610 15.90179
#> 16 27.69537 27.73132 27.76476 27.79602 27.82519 27.85238 27.87775 27.90140
#> 17 27.03186 27.01943 27.00795 26.99713 26.98704 26.97763 26.96886 26.96067
#> 18 28.80235 28.79574 28.78951 28.78376 28.77839 28.77339 28.76873 28.76438
#> 19 21.51412 21.36637 21.22893 21.10043 20.98056 20.86878 20.76452 20.66730
#> 20 15.68560 15.49392 15.31553 15.14881 14.99330 14.84827 14.71302 14.58688
#> 21 25.22467 25.20630 25.18944 25.17346 25.15855 25.14464 25.13167 25.11957
#>      0.3000   0.2798   0.2609   0.2433   0.2269   0.2116   0.1974   0.1841
#> 1  22.19278 22.19287 22.19295 22.19301 21.86398 21.86412 21.95477 21.95475
#> 2  21.49508 21.49513 21.49517 21.49521 21.10914 21.10922 21.38570 21.38579
#> 3  26.52988 26.52979 26.52971 26.52964 26.51212 26.51205 27.11213 27.11242
#> 4  20.36327 20.36320 20.36314 20.36309 20.72093 20.72092 20.63304 20.63297
#> 5  18.66593 18.66567 18.66545 18.66527 18.68998 18.69003 18.99917 18.99919
#> 6  16.13244 16.13258 16.13270 16.13280 16.23640 16.23656 15.62473 15.62436
#> 7  23.67354 23.67323 23.67298 23.67277 23.56027 23.55991 22.69195 22.69181
#> 8  20.01338 20.01335 20.01332 20.01330 21.14076 21.14017 20.16508 20.16491
#> 9  15.91860 15.91875 15.91888 15.91898 15.82405 15.82430 15.62031 15.62012
#> 10 11.34006 11.33994 11.33983 11.33974 11.06370 11.06347 11.73147 11.73196
#> 11 27.59637 27.59640 27.59643 27.59645 27.80582 27.80570 28.16129 28.16155
#> 12 29.04436 29.04451 29.04464 29.04475 29.45752 29.45743 29.45146 29.45156
#> 13 16.44512 16.44528 16.44541 16.44551 16.10933 16.10982 16.08724 16.08699
#> 14 17.92935 17.92970 17.92999 17.93024 17.95227 17.95264 17.60185 17.60160
#> 15 16.58289 16.58315 16.58336 16.58353 16.45298 16.45329 16.44907 16.44900
#> 16 28.32143 28.32151 28.32157 28.32162 28.59026 28.59020 28.75268 28.75281
#> 17 28.07377 28.07387 28.07395 28.07402 27.00850 27.00872 26.70620 26.70615
#> 18 27.67111 27.67099 27.67090 27.67082 27.79402 27.79407 27.70145 27.70128
#> 19 19.26364 19.26336 19.26313 19.26294 18.71640 18.71643 18.78911 18.78896
#> 20 13.95133 13.95120 13.95110 13.95101 14.46078 14.46049 14.85231 14.85237
#> 21 25.29567 25.29552 25.29539 25.29528 25.43079 25.43045 26.02900 26.02945
#>      0.1717   0.1601   0.1493   0.1392   0.1299   0.1211   0.1129   0.1053
#> 1  21.95473 22.12728 22.12726 22.12725 22.12725 22.12725 22.12725 22.46400
#> 2  21.38587 21.45638 21.45643 21.45647 21.45652 21.45657 21.45661 21.92437
#> 3  27.11265 26.15024 26.15038 26.15051 26.15063 26.15073 26.15081 25.47422
#> 4  20.63292 21.79047 21.79047 21.79046 21.79045 21.79042 21.79040 21.82617
#> 5  18.99921 18.76896 18.76907 18.76918 18.76926 18.76934 18.76941 18.47073
#> 6  15.62409 14.58397 14.58393 14.58390 14.58388 14.58386 14.58386 13.36556
#> 7  22.69171 24.63570 24.63575 24.63579 24.63583 24.63586 24.63590 24.39457
#> 8  20.16479 18.45184 18.45096 18.45020 18.44955 18.44900 18.44853 19.31859
#> 9  15.61998 13.15805 13.15806 13.15808 13.15811 13.15815 13.15820 13.11905
#> 10 11.73232 12.42408 12.42417 12.42424 12.42430 12.42435 12.42438 11.98381
#> 11 28.16174 27.76957 27.76969 27.76980 27.76989 27.76998 27.77006 28.06900
#> 12 29.45162 30.15173 30.15196 30.15216 30.15233 30.15248 30.15261 30.96208
#> 13 16.08681 15.75181 15.75177 15.75173 15.75169 15.75165 15.75162 15.81279
#> 14 17.60140 17.06549 17.06538 17.06529 17.06522 17.06515 17.06509 17.73063
#> 15 16.44895 18.97456 18.97468 18.97478 18.97484 18.97487 18.97488 19.57704
#> 16 28.75291 28.49719 28.49724 28.49730 28.49734 28.49738 28.49742 28.66822
#> 17 26.70612 27.03818 27.03791 27.03766 27.03744 27.03725 27.03707 26.13710
#> 18 27.70114 28.76018 28.76012 28.76005 28.75998 28.75991 28.75983 28.56170
#> 19 18.78885 18.31780 18.31771 18.31764 18.31759 18.31755 18.31752 17.94512
#> 20 14.85241 15.90125 15.90172 15.90214 15.90250 15.90282 15.90309 16.74094
#> 21 26.02979 24.72529 24.72532 24.72536 24.72540 24.72543 24.72547 23.95434
#>      0.0982   0.0916   0.0854   0.0797   0.0743   0.0693   0.0646   0.0603
#> 1  22.46395 22.46391 22.46387 22.46383 22.46380 22.46377 22.46374 22.46371
#> 2  21.92431 21.92425 21.92420 21.92415 21.92410 21.92406 21.92402 21.92398
#> 3  25.47435 25.47448 25.47460 25.47472 25.47484 25.47495 25.47505 25.47515
#> 4  21.82614 21.82611 21.82607 21.82603 21.82599 21.82595 21.82590 21.82586
#> 5  18.47080 18.47087 18.47093 18.47100 18.47105 18.47111 18.47116 18.47121
#> 6  13.36578 13.36600 13.36622 13.36643 13.36664 13.36685 13.36705 13.36725
#> 7  24.39462 24.39467 24.39472 24.39477 24.39481 24.39486 24.39490 24.39495
#> 8  19.31838 19.31819 19.31803 19.31788 19.31776 19.31765 19.31755 19.31747
#> 9  13.11911 13.11917 13.11924 13.11931 13.11939 13.11947 13.11955 13.11964
#> 10 11.98387 11.98393 11.98398 11.98402 11.98406 11.98409 11.98412 11.98415
#> 11 28.06896 28.06893 28.06889 28.06886 28.06882 28.06879 28.06875 28.06872
#> 12 30.96195 30.96183 30.96170 30.96158 30.96146 30.96134 30.96122 30.96110
#> 13 15.81277 15.81276 15.81274 15.81272 15.81271 15.81269 15.81267 15.81265
#> 14 17.73051 17.73039 17.73028 17.73017 17.73006 17.72996 17.72986 17.72977
#> 15 19.57689 19.57673 19.57657 19.57640 19.57623 19.57605 19.57587 19.57568
#> 16 28.66820 28.66817 28.66815 28.66812 28.66810 28.66808 28.66805 28.66803
#> 17 26.13721 26.13732 26.13743 26.13754 26.13764 26.13774 26.13783 26.13792
#> 18 28.56169 28.56168 28.56167 28.56165 28.56162 28.56159 28.56156 28.56152
#> 19 17.94520 17.94528 17.94537 17.94546 17.94555 17.94564 17.94573 17.94582
#> 20 16.74082 16.74070 16.74058 16.74045 16.74033 16.74020 16.74007 16.73994
#> 21 23.95448 23.95463 23.95477 23.95492 23.95506 23.95519 23.95533 23.95546
#>      0.0562   0.0524   0.0489   0.0456   0.0425   0.0397   0.0370   0.0345
#> 1  22.46369 22.46367 22.46365 22.46364 22.46362 22.46361 22.12842 22.12835
#> 2  21.92395 21.92392 21.92389 21.92386 21.92384 21.92382 21.70095 21.70093
#> 3  25.47525 25.47534 25.47543 25.47552 25.47560 25.47567 25.63753 25.63750
#> 4  21.82581 21.82576 21.82571 21.82565 21.82560 21.82555 21.59824 21.59823
#> 5  18.47126 18.47131 18.47135 18.47139 18.47143 18.47147 18.45799 18.45796
#> 6  13.36744 13.36763 13.36781 13.36799 13.36817 13.36834 13.13934 13.13906
#> 7  24.39499 24.39503 24.39507 24.39511 24.39515 24.39518 24.22947 24.22938
#> 8  19.31740 19.31735 19.31731 19.31728 19.31726 19.31725 19.32120 19.32124
#> 9  13.11973 13.11982 13.11992 13.12001 13.12011 13.12021 13.43546 13.43545
#> 10 11.98417 11.98418 11.98420 11.98421 11.98421 11.98422 12.02956 12.02957
#> 11 28.06869 28.06866 28.06863 28.06860 28.06857 28.06854 28.20581 28.20589
#> 12 30.96099 30.96088 30.96077 30.96066 30.96055 30.96045 30.95817 30.95828
#> 13 15.81263 15.81261 15.81259 15.81257 15.81255 15.81253 15.91632 15.91637
#> 14 17.72968 17.72960 17.72952 17.72944 17.72936 17.72929 17.94276 17.94289
#> 15 19.57550 19.57531 19.57513 19.57494 19.57475 19.57457 19.53124 19.53140
#> 16 28.66801 28.66799 28.66797 28.66795 28.66793 28.66791 28.68601 28.68603
#> 17 26.13801 26.13810 26.13818 26.13826 26.13833 26.13841 26.38549 26.38547
#> 18 28.56149 28.56145 28.56140 28.56136 28.56131 28.56127 28.57425 28.57430
#> 19 17.94591 17.94600 17.94610 17.94619 17.94628 17.94637 17.78398 17.78386
#> 20 16.73981 16.73968 16.73955 16.73943 16.73930 16.73917 16.80669 16.80685
#> 21 23.95559 23.95571 23.95584 23.95595 23.95607 23.95618 24.03109 24.03099
#>      0.0322   0.0300   0.0280   0.0261   0.0243   0.0227   0.0212   0.0197
#> 1  22.12829 22.12822 22.12816 22.12811 22.12805 22.12800 22.12796 22.12791
#> 2  21.70091 21.70089 21.70088 21.70086 21.70085 21.70083 21.70082 21.70081
#> 3  25.63746 25.63742 25.63739 25.63735 25.63732 25.63728 25.63725 25.63721
#> 4  21.59822 21.59821 21.59820 21.59820 21.59820 21.59820 21.59820 21.59820
#> 5  18.45794 18.45792 18.45789 18.45787 18.45785 18.45783 18.45781 18.45779
#> 6  13.13878 13.13851 13.13825 13.13800 13.13776 13.13753 13.13730 13.13708
#> 7  24.22930 24.22922 24.22914 24.22906 24.22899 24.22891 24.22885 24.22878
#> 8  19.32127 19.32130 19.32133 19.32136 19.32139 19.32141 19.32143 19.32144
#> 9  13.43543 13.43541 13.43539 13.43537 13.43535 13.43532 13.43530 13.43527
#> 10 12.02958 12.02959 12.02960 12.02962 12.02963 12.02964 12.02965 12.02966
#> 11 28.20596 28.20602 28.20609 28.20615 28.20621 28.20627 28.20632 28.20638
#> 12 30.95838 30.95847 30.95857 30.95866 30.95875 30.95884 30.95892 30.95900
#> 13 15.91642 15.91647 15.91651 15.91656 15.91660 15.91664 15.91667 15.91671
#> 14 17.94301 17.94313 17.94324 17.94335 17.94346 17.94356 17.94365 17.94375
#> 15 19.53157 19.53173 19.53188 19.53203 19.53219 19.53233 19.53248 19.53262
#> 16 28.68605 28.68608 28.68610 28.68612 28.68614 28.68616 28.68617 28.68619
#> 17 26.38544 26.38542 26.38539 26.38537 26.38534 26.38531 26.38529 26.38526
#> 18 28.57435 28.57439 28.57444 28.57448 28.57452 28.57457 28.57461 28.57465
#> 19 17.78375 17.78365 17.78354 17.78344 17.78334 17.78325 17.78315 17.78306
#> 20 16.80700 16.80714 16.80728 16.80742 16.80755 16.80768 16.80781 16.80793
#> 21 24.03090 24.03080 24.03071 24.03063 24.03054 24.03046 24.03037 24.03029
#>      0.0184   0.0172   0.0160   0.0149   0.0139   0.0130   0.0121   0.0113
#> 1  22.12787 22.12783 22.12779 22.12775 22.12772 22.12769 22.12766 22.12763
#> 2  21.70080 21.70079 21.70079 21.70078 21.70077 21.70077 21.70076 21.70076
#> 3  25.63718 25.63715 25.63711 25.63708 25.63705 25.63702 25.63699 25.63696
#> 4  21.59820 21.59821 21.59822 21.59822 21.59823 21.59824 21.59825 21.59826
#> 5  18.45777 18.45775 18.45774 18.45772 18.45770 18.45769 18.45767 18.45766
#> 6  13.13687 13.13666 13.13646 13.13627 13.13609 13.13591 13.13573 13.13557
#> 7  24.22872 24.22866 24.22860 24.22854 24.22849 24.22844 24.22839 24.22834
#> 8  19.32146 19.32147 19.32148 19.32149 19.32150 19.32150 19.32151 19.32151
#> 9  13.43524 13.43521 13.43518 13.43515 13.43511 13.43508 13.43505 13.43501
#> 10 12.02967 12.02968 12.02969 12.02970 12.02971 12.02972 12.02973 12.02974
#> 11 28.20643 28.20648 28.20652 28.20657 28.20661 28.20665 28.20669 28.20673
#> 12 30.95908 30.95916 30.95924 30.95931 30.95939 30.95946 30.95952 30.95959
#> 13 15.91675 15.91678 15.91681 15.91685 15.91688 15.91690 15.91693 15.91696
#> 14 17.94383 17.94392 17.94400 17.94408 17.94415 17.94423 17.94429 17.94436
#> 15 19.53276 19.53289 19.53303 19.53316 19.53329 19.53341 19.53353 19.53365
#> 16 28.68621 28.68623 28.68624 28.68626 28.68627 28.68629 28.68630 28.68632
#> 17 26.38523 26.38520 26.38518 26.38515 26.38512 26.38509 26.38507 26.38504
#> 18 28.57469 28.57473 28.57477 28.57481 28.57484 28.57488 28.57492 28.57495
#> 19 17.78298 17.78289 17.78281 17.78273 17.78265 17.78258 17.78251 17.78244
#> 20 16.80805 16.80817 16.80828 16.80839 16.80850 16.80861 16.80871 16.80881
#> 21 24.03021 24.03014 24.03006 24.02999 24.02992 24.02985 24.02978 24.02971
#>      0.0105   0.0098   0.0092   0.0085   0.0080   0.0074   0.0069   0.0065
#> 1  22.12760 22.12758 22.12756 22.12753 22.12751 22.12750 22.12748 22.12746
#> 2  21.70076 21.70075 21.70075 21.70075 21.70075 21.70075 21.70075 21.70075
#> 3  25.63693 25.63690 25.63687 25.63684 25.63682 25.63679 25.63676 25.63674
#> 4  21.59827 21.59829 21.59830 21.59831 21.59833 21.59834 21.59836 21.59838
#> 5  18.45765 18.45763 18.45762 18.45761 18.45760 18.45759 18.45758 18.45757
#> 6  13.13541 13.13525 13.13510 13.13496 13.13482 13.13469 13.13456 13.13443
#> 7  24.22830 24.22826 24.22822 24.22818 24.22814 24.22810 24.22807 24.22804
#> 8  19.32151 19.32151 19.32151 19.32150 19.32150 19.32149 19.32149 19.32148
#> 9  13.43497 13.43494 13.43490 13.43486 13.43482 13.43479 13.43475 13.43471
#> 10 12.02974 12.02975 12.02976 12.02977 12.02978 12.02979 12.02980 12.02981
#> 11 28.20677 28.20680 28.20683 28.20687 28.20690 28.20693 28.20695 28.20698
#> 12 30.95966 30.95972 30.95978 30.95984 30.95990 30.95995 30.96001 30.96006
#> 13 15.91699 15.91701 15.91703 15.91706 15.91708 15.91710 15.91712 15.91714
#> 14 17.94442 17.94448 17.94454 17.94460 17.94465 17.94470 17.94474 17.94479
#> 15 19.53377 19.53389 19.53400 19.53411 19.53421 19.53432 19.53442 19.53452
#> 16 28.68633 28.68634 28.68636 28.68637 28.68638 28.68639 28.68640 28.68641
#> 17 26.38501 26.38498 26.38496 26.38493 26.38490 26.38488 26.38485 26.38482
#> 18 28.57499 28.57502 28.57506 28.57509 28.57512 28.57515 28.57518 28.57521
#> 19 17.78237 17.78231 17.78225 17.78218 17.78213 17.78207 17.78202 17.78196
#> 20 16.80890 16.80899 16.80908 16.80917 16.80926 16.80934 16.80942 16.80950
#> 21 24.02965 24.02959 24.02952 24.02946 24.02941 24.02935 24.02929 24.02924
#>      0.0060   0.0056   0.0052   0.0049
#> 1  22.12745 22.12744 22.12743 22.12741
#> 2  21.70075 21.70075 21.70075 21.70075
#> 3  25.63671 25.63669 25.63666 25.63664
#> 4  21.59839 21.59841 21.59843 21.59845
#> 5  18.45756 18.45755 18.45754 18.45753
#> 6  13.13432 13.13420 13.13409 13.13399
#> 7  24.22801 24.22798 24.22795 24.22792
#> 8  19.32147 19.32146 19.32145 19.32144
#> 9  13.43467 13.43463 13.43459 13.43455
#> 10 12.02982 12.02983 12.02984 12.02984
#> 11 28.20700 28.20703 28.20705 28.20707
#> 12 30.96011 30.96016 30.96021 30.96026
#> 13 15.91715 15.91717 15.91719 15.91720
#> 14 17.94483 17.94487 17.94491 17.94495
#> 15 19.53462 19.53471 19.53481 19.53490
#> 16 28.68642 28.68643 28.68644 28.68645
#> 17 26.38480 26.38477 26.38475 26.38472
#> 18 28.57524 28.57527 28.57530 28.57533
#> 19 17.78191 17.78186 17.78182 17.78177
#> 20 16.80957 16.80965 16.80972 16.80978
#> 21 24.02919 24.02914 24.02909 24.02904
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 22.8 21.4 18.1 14.3 22.8 19.2 15.2 14.7 32.4 33.9 15.5 15.2 19.2
#> [16] 27.3 26.0 30.4 19.7 15.0 21.4
#> 
#> $X
#>               am       carb         cyl       disp       drat       gear
#>  [1,]  0.9534626  0.7129994  0.05643326 -0.4177430  0.6071012  0.2603778
#>  [2,]  0.9534626  0.7129994  0.05643326 -0.4177430  0.6071012  0.2603778
#>  [3,]  0.9534626 -0.9506659 -1.12866530 -0.8888039  0.5046163  0.2603778
#>  [4,] -1.0488088 -0.9506659  0.05643326  0.4700256 -1.0736517 -1.1066057
#>  [5,] -1.0488088 -0.9506659  0.05643326  0.1710831 -1.7295552 -1.1066057
#>  [6,] -1.0488088  0.7129994  1.24153183  1.3940296 -0.8071908 -1.1066057
#>  [7,] -1.0488088 -0.3961108 -1.12866530 -0.5916732  0.6480952  0.2603778
#>  [8,] -1.0488088  0.7129994  0.05643326 -0.3488956  0.6480952  0.2603778
#>  [9,] -1.0488088  0.1584443  1.24153183  0.6312733 -1.0941486 -1.1066057
#> [10,] -1.0488088  0.7129994  1.24153183  2.1187386 -0.7661969 -1.1066057
#> [11,]  0.9534626 -0.9506659 -1.12866530 -1.1542285  0.9760470  0.2603778
#> [12,]  0.9534626 -0.9506659 -1.12866530 -1.2230759  1.2630048  0.2603778
#> [13,] -1.0488088 -0.3961108  1.24153183  1.0135573 -1.7295552 -1.1066057
#> [14,] -1.0488088 -0.3961108  1.24153183  0.8867333 -0.9301727 -1.1066057
#> [15,] -1.0488088 -0.3961108  1.24153183  1.7563841 -1.0736517 -1.1066057
#> [16,]  0.9534626 -0.9506659 -1.12866530 -1.1515109  0.9760470  0.2603778
#> [17,]  0.9534626 -0.3961108 -1.12866530 -0.7773798  1.6934415  1.6273614
#> [18,]  0.9534626 -0.3961108 -1.12866530 -1.0056632  0.3406404  1.6273614
#> [19,]  0.9534626  1.8221097  0.05643326 -0.5536259  0.0331856  1.6273614
#> [20,]  0.9534626  2.9312199  1.24153183  0.8595567 -0.1307903  1.6273614
#> [21,]  0.9534626 -0.3961108 -1.12866530 -0.7710386  1.0375379  0.2603778
#>               hp       qsec         vs         wt
#>  [1,] -0.4252522 -0.7916798 -0.9534626 -0.4643627
#>  [2,] -0.4252522 -0.4848191 -0.9534626 -0.1646926
#>  [3,] -0.6843222  0.3864462  1.0488088 -0.8169157
#>  [4,] -0.4252522  0.8412576  1.0488088  0.2348675
#>  [5,] -0.5014493  1.2686708  1.0488088  0.5227858
#>  [6,]  1.6320686 -1.1314185 -0.9534626  0.6520552
#>  [7,] -0.6538434  2.7372185  1.0488088  0.1584810
#>  [8,] -0.2271398  0.2165769  1.0488088  0.4992822
#>  [9,]  0.6415067  0.0521872 -0.9534626  0.8988423
#> [10,]  1.4034774 -0.2656328 -0.9534626  2.7379938
#> [11,] -1.0957864  0.8576966  1.0488088 -0.9579369
#> [12,] -1.1110258  1.0933218  1.0488088 -1.3868764
#> [13,]  0.1843243 -0.5670139 -0.9534626  0.5932964
#> [14,]  0.1843243 -0.3313887 -0.9534626  0.4934063
#> [15,]  0.5653097 -0.4683801 -0.9534626  0.9752288
#> [16,] -1.0957864  0.5453562  1.0488088 -1.2693587
#> [17,] -0.7148011 -0.6601681 -0.9534626 -1.0284475
#> [18,] -0.3795340 -0.5505749  1.0488088 -1.7652833
#> [19,]  0.5653097 -1.3177268 -0.9534626 -0.2880862
#> [20,]  3.0036158 -1.8108958 -0.9534626  0.6520552
#> [21,] -0.4404916  0.3809665  1.0488088 -0.2763344
#> attr(,"center")
#>  [1]   0.5238095   2.7142857   5.9047619 206.1142857   3.6038095   3.8095238
#>  [7] 137.9047619  17.9047619   0.4761905   3.0151429
#> attr(,"scale")
#>  [1]   0.4994328   1.8032473   1.6876233 110.3891311   0.4878766   0.7315377
#>  [7]  65.6193241   1.8249321   0.4994328   0.8509359
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 30
#> 
#> $lambda.min
#> [1] 0.6463328
#> 
#> $null.dev
#> [1] 32.83188
#> 
#> $Bias
#> [1] 8.855143
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 4.336268