Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 41.91775 38.73253 35.71801 33.34426 30.07463 27.33620 25.12522 23.36149
#>   [9] 21.97600 20.90949 20.11108 19.53712 19.15029 18.91861 18.81483 18.80062
#>  [17] 18.46823 18.10447 17.67752 17.23715 17.05548 16.89567 16.75500 16.32245
#>  [25] 15.25856 14.56733 14.28748 14.19785 14.17711 14.20529 14.38036 14.47605
#>  [33] 14.50818 14.43537 14.38736 14.39624 14.06699 14.04359 14.03338 14.31181
#>  [41] 14.32264 14.41366 12.82452 12.89121 13.24979 12.67041 11.04906 11.88595
#>  [49] 11.47200 11.95389 11.94846 11.59067 11.91460 30.18450 29.48176 29.62530
#>  [57] 29.62760 29.39523 28.97027 28.69667 28.72671 28.32742 27.90520 27.90719
#>  [65] 27.90916 27.91110 27.91303 27.91493 27.86352 27.86548 27.90662 27.90847
#>  [73] 28.08000 27.93468 27.93642 27.93814 27.93980 27.85091 27.85246 27.91018
#>  [81] 27.91164 27.91309 27.91452 27.91595 27.86877 27.77149 27.76812 27.76810
#>  [89] 27.76905 27.77001 27.77098 27.77196 27.77295 27.77394 27.77509 28.47208
#>  [97] 28.47340 28.47472 28.47604 28.47729
#> 
#> $cvse
#>   [1] 12.847945 12.569070 12.294444 12.096998 11.035926 10.033061  9.131071
#>   [8]  8.320123  7.593298  6.946061  6.375781  5.881234  5.462052  5.118068
#>  [15]  4.848579  4.647070  4.406198  4.196658  4.008591  3.871757  3.868240
#>  [22]  3.866601  3.866201  3.831127  3.812063  3.844895  3.861040  3.862949
#>  [29]  3.852371  3.818357  3.779335  3.811074  3.829996  3.816258  3.809898
#>  [36]  3.803186  3.837239  3.818264  3.800262  3.870634  3.861799  3.854186
#>  [43]  3.928103  3.917730  3.913320  3.767635  3.102424  3.240441  3.107880
#>  [50]  3.029678  2.958366  2.690343  2.737980 15.445071 15.488549 15.481534
#>  [57] 15.483300 15.498285 15.527664 15.547420 15.552490 15.551803 15.566754
#>  [64] 15.568386 15.570001 15.571601 15.573186 15.574755 15.578964 15.580498
#>  [71] 15.580633 15.582138 15.573881 15.579564 15.581026 15.582473 15.583908
#>  [78] 15.590895 15.592310 15.591110 15.592500 15.593877 15.595239 15.596589
#>  [85] 15.595800 15.604855 15.604016 15.605573 15.606880 15.608174 15.609455
#>  [92] 15.610724 15.611980 15.613223 15.614443 15.575117 15.576312 15.577495
#>  [99] 15.578666 15.579823
#> 
#> $fold
#>  [1]  9  5  1  4  2  1  4  7  6  3 10  7  8  1  3 10  5  8  2  6  9
#> 
#> $lambda
#>   [1] 5.313814072 4.955680788 4.621684489 4.310198423 4.019705475 3.748790779
#>   [7] 3.496134827 3.260507041 3.040759779 2.835822747 2.644697785 2.466454006
#>  [13] 2.300223261 2.145195912 2.000616887 1.865782005 1.740034542 1.622762038
#>  [19] 1.513393307 1.411395664 1.316272320 1.227559971 1.144826537 1.067669060
#>  [25] 0.995711739 0.928604100 0.866019292 0.807652490 0.753219415 0.702454948
#>  [31] 0.655111835 0.610959490 0.569782865 0.531381407 0.495568079 0.462168449
#>  [37] 0.431019842 0.401970547 0.374879078 0.349613483 0.326050704 0.304075978
#>  [43] 0.283582275 0.264469779 0.246645401 0.230022326 0.214519591 0.200061689
#>  [49] 0.186578200 0.174003454 0.162276204 0.151339331 0.141139566 0.131627232
#>  [55] 0.122755997 0.114482654 0.106766906 0.099571174 0.092860410 0.086601929
#>  [61] 0.080765249 0.075321942 0.070245495 0.065511184 0.061095949 0.056978287
#>  [67] 0.053138141 0.049556808 0.046216845 0.043101984 0.040197055 0.037487908
#>  [73] 0.034961348 0.032605070 0.030407598 0.028358227 0.026446978 0.024664540
#>  [79] 0.023002233 0.021451959 0.020006169 0.018657820 0.017400345 0.016227620
#>  [85] 0.015133933 0.014113957 0.013162723 0.012275600 0.011448265 0.010676691
#>  [91] 0.009957117 0.009286041 0.008660193 0.008076525 0.007532194 0.007024549
#>  [97] 0.006551118 0.006109595 0.005697829 0.005313814
#> 
#> $fit
#> $beta
#>               5.3138       4.9557       4.6217      4.3102      4.0197
#> (Intercept) 19.24286 20.260659036 21.209864487 22.09509667 22.92066716
#> am           0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> carb         0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> cyl          0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> disp         0.00000 -0.004197286 -0.008111689 -0.01176227 -0.01516682
#> drat         0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> gear         0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> hp           0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> qsec         0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> vs           0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#> wt           0.00000  0.000000000  0.000000000  0.00000000  0.00000000
#>                  3.7488      3.4961      3.2605      3.0408      2.8358
#> (Intercept) 23.69059697 24.40863608 25.07828177 25.70279557 26.28521923
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.01834192 -0.02130302 -0.02406455 -0.02663997 -0.02904181
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  2.6447      2.4665      2.3002      2.1452      2.0006
#> (Intercept) 26.82838949 27.33495189 27.80737368 28.24795582 28.65884421
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.03128177 -0.03337077 -0.03531898 -0.03713589 -0.03883034
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  1.8658      1.7400      1.6228      1.5134      1.4114
#> (Intercept) 29.04204008 29.31063594 29.31063594 29.31063594 29.31063594
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.04041059 -0.04151824 -0.04151824 -0.04151824 -0.04151824
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  1.3163      1.2276      1.1448      1.0677       0.9957
#> (Intercept) 29.31063594 29.31063594 29.31063594 29.31063594 29.321325087
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.000000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000 -0.005940867
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.000000000
#> disp        -0.04151824 -0.04151824 -0.04151824 -0.04151824 -0.041485327
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.000000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.000000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.000000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.000000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.000000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.000000000
#>                  0.9286      0.8660      0.8077      0.7532      0.7025
#> (Intercept) 29.46997942 29.60862917 29.73793441 29.85852491 29.97098802
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb        -0.08856071 -0.16562016 -0.23748607 -0.30450846 -0.36701376
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.04102754 -0.04060057 -0.04020237 -0.03983101 -0.03948468
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.6551      0.6110      0.5698      0.5314      0.4956
#> (Intercept) 30.07587148 29.41150573 27.87168346 27.87125192 27.87097228
#> am           0.00000000  1.20832442  3.55701690  3.55769527  3.55813485
#> carb        -0.42530642 -0.71685775 -1.14547606 -1.14560719 -1.14569216
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.03916169 -0.03454149 -0.02632604 -0.02632362 -0.02632206
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.4622      0.4310    0.4020    0.3749    0.3496    0.3261
#> (Intercept) 27.87079107 27.87067365 -4.127471 -4.126460 -4.125889 -4.125568
#> am           3.55841969  3.55860427  0.000000  0.000000  0.000000  0.000000
#> carb        -1.14574722 -1.14578290 -2.273551 -2.273654 -2.273712 -2.273745
#> cyl          0.00000000  0.00000000  0.000000  0.000000  0.000000  0.000000
#> disp        -0.02632104 -0.02632039  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000000  0.00000000  3.661123  3.660701  3.660464  3.660330
#> gear         0.00000000  0.00000000  4.832529  4.832753  4.832879  4.832951
#> hp           0.00000000  0.00000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000000  0.00000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000000  0.00000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000000  0.00000000  0.000000  0.000000  0.000000  0.000000
#>                0.3041    0.2836    0.2645    0.2466    0.2300    0.2145
#> (Intercept) -4.125386 -4.125284 -4.125226  4.902045  4.905121  4.907534
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb        -2.273763 -2.273774 -2.273779 -1.900884 -1.900638 -1.900446
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         3.660254  3.660211  3.660187  2.864921  2.864892  2.864869
#> gear         4.832991  4.833014  4.833026  3.872728  3.872095  3.871599
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000 -1.157979 -1.158412 -1.158752
#>                0.2001    0.1866    0.1740    0.1623     0.1513     0.1411
#> (Intercept)  4.909425  4.910908  4.912070  4.912982 -30.703924 -30.710052
#> am           0.000000  0.000000  0.000000  0.000000   0.000000   0.000000
#> carb        -1.900295 -1.900177 -1.900084 -1.900012  -1.935798  -1.935984
#> cyl          0.000000  0.000000  0.000000  0.000000   1.434894   1.435201
#> disp         0.000000  0.000000  0.000000  0.000000   0.000000   0.000000
#> drat         2.864852  2.864838  2.864827  2.864818   3.858165   3.858419
#> gear         3.871211  3.870906  3.870666  3.870479   5.319587   5.320111
#> hp           0.000000  0.000000  0.000000  0.000000   0.000000   0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000   1.140685   1.140789
#> vs           0.000000  0.000000  0.000000  0.000000   0.000000   0.000000
#> wt          -1.159018 -1.159226 -1.159390 -1.159518  -1.833413  -1.833374
#>                 0.1316     0.1228     0.1145     0.1068       0.0996
#> (Intercept) -30.715715 -34.176455 -34.180262 -34.183718 -71.95068586
#> am            0.000000   0.000000   0.000000   0.000000  -3.39730712
#> carb         -1.936156  -2.058610  -2.058735  -2.058849  -3.31172463
#> cyl           1.435485   1.143880   1.144177   1.144448   0.69615791
#> disp          0.000000   0.000000   0.000000   0.000000   0.00000000
#> drat          3.858653   4.315910   4.315993   4.316065   7.66291751
#> gear          5.320596   5.627094   5.627434   5.627743   9.01201263
#> hp            0.000000   0.000000   0.000000   0.000000   0.03454154
#> qsec          1.140884   1.339549   1.339587   1.339622   2.37811184
#> vs            0.000000  -2.270355  -2.269956  -2.269586  -7.69028048
#> wt           -1.833338  -1.731888  -1.731901  -1.731915  -1.72871607
#>                   0.0929       0.0866       0.0808       0.0753     0.0702
#> (Intercept) -71.95942411 -71.96798723 -71.97637873 -71.98460206 -71.992661
#> am           -3.39794283  -3.39856580  -3.39917628  -3.39977453  -3.400361
#> carb         -3.31203058  -3.31233040  -3.31262421  -3.31291214  -3.313194
#> cyl           0.69614253   0.69612746   0.69611269   0.69609822   0.696084
#> disp          0.00000000   0.00000000   0.00000000   0.00000000   0.000000
#> drat          7.66367103   7.66440945   7.66513306   7.66584218   7.666537
#> gear          9.01283417   9.01363925   9.01442818   9.01520131   9.015959
#> hp            0.03454746   0.03455327   0.03455896   0.03456453   0.034570
#> qsec          2.37831674   2.37851753   2.37871430   2.37890712   2.379096
#> vs           -7.69137877  -7.69245504  -7.69350975  -7.69454332  -7.695556
#> wt           -1.72863922  -1.72856391  -1.72849011  -1.72841778  -1.728347
#>                   0.0655      0.0611       0.0570       0.0531       0.0496
#> (Intercept) -72.00055761 -72.0082964 -72.01588004 -72.02331172 -72.03059446
#> am           -3.40093531  -3.4014983  -3.40205002  -3.40259067  -3.40312050
#> carb         -3.31347079  -3.3137417  -3.31400727  -3.31426747  -3.31452246
#> cyl           0.69607014   0.6960565   0.69604318   0.69603010   0.69601728
#> disp          0.00000000   0.0000000   0.00000000   0.00000000   0.00000000
#> drat          7.66721805   7.6678854   7.66853934   7.66918019   7.66980819
#> gear          9.01670140   9.0174290   9.01814195   9.01884065   9.01952535
#> hp            0.03457535   0.0345806   0.03458574   0.03459078   0.03459572
#> qsec          2.37928125   2.3794627   2.37964053   2.37981479   2.37998555
#> vs           -7.69654873  -7.6975214  -7.69847456  -7.69940863  -7.70032397
#> wt           -1.72827745  -1.7282094  -1.72814269  -1.72807733  -1.72801328
#>                   0.0462      0.0431       0.0402      0.0375       0.0350
#> (Intercept) -72.03773125 -72.0447250 -72.05157860 -72.0582948 -72.06487649
#> am           -3.40363970  -3.4041485  -3.40464710  -3.4051357  -3.40561452
#> carb         -3.31477234  -3.3150172  -3.31525718  -3.3154923  -3.31572277
#> cyl           0.69600472   0.6959924   0.69598035   0.6959685   0.69595695
#> disp          0.00000000   0.0000000   0.00000000   0.0000000   0.00000000
#> drat          7.67042361   7.6710267   7.67161770   7.6721969   7.67276440
#> gear          9.02019633   9.0208539   9.02149821   9.0221296   9.02274842
#> hp            0.03460056   0.0346053   0.03460995   0.0346145   0.03461896
#> qsec          2.38015290   2.3803169   2.38047759   2.3806351   2.38078940
#> vs           -7.70122098  -7.7021000  -7.70296141  -7.7038056  -7.70463279
#> wt           -1.72795052  -1.7278890  -1.72782873  -1.7277697  -1.72771178
#>                   0.0326       0.0304       0.0284       0.0264       0.0247
#> (Intercept) -72.07132622 -72.07764670 -72.08384051 -72.08991019 -72.09585822
#> am           -3.40608374  -3.40654356  -3.40699416  -3.40743573  -3.40786845
#> carb         -3.31594860  -3.31616990  -3.31638676  -3.31659928  -3.31680753
#> cyl           0.69594560   0.69593448   0.69592358   0.69591290   0.69590243
#> disp          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> drat          7.67332057   7.67386560   7.67439970   7.67492310   7.67543601
#> gear          9.02335481   9.02394903   9.02453135   9.02510200   9.02566122
#> hp            0.03462334   0.03462762   0.03463182   0.03463594   0.03463997
#> qsec          2.38094064   2.38108884   2.38123408   2.38137640   2.38151587
#> vs           -7.70544344  -7.70623784  -7.70701632  -7.70777920  -7.70852679
#> wt           -1.72765505  -1.72759946  -1.72754499  -1.72749161  -1.72743929
#>                   0.0230      0.0215       0.0200       0.0187        0.0174
#> (Intercept) -72.10168705 -72.1073991 -72.11299660 -72.11848196 -72.554195462
#> am           -3.40829250  -3.4087081  -3.40911528  -3.40951434  -3.354325169
#> carb         -3.31701162  -3.3172116  -3.31740760  -3.31759966  -3.221983511
#> cyl           0.69589217   0.6958821   0.69587227   0.69586262   0.684240678
#> disp          0.00000000   0.0000000   0.00000000   0.00000000   0.002946316
#> drat          7.67593865   7.6764312   7.67691389   7.67738690   7.598648218
#> gear          9.02620922   9.0267462   9.02727250   9.02778822   8.999989259
#> hp            0.03464392   0.0346478   0.03465159   0.03465531   0.032787930
#> qsec          2.38165254   2.3817865   2.38191773   2.38204635   2.440132225
#> vs           -7.70925940  -7.7099773  -7.71068087  -7.71137031  -7.617351867
#> wt           -1.72738803  -1.7273378  -1.72728856  -1.72724032  -1.996556899
#>                    0.0162        0.0151        0.0141        0.0132
#> (Intercept) -72.545799341 -72.537629259 -72.529679582 -72.521944707
#> am           -3.353836520  -3.353360265  -3.352895978  -3.352443246
#> carb         -3.221657235  -3.221341051  -3.221034590  -3.220737495
#> cyl           0.684111210   0.683988529   0.683872364   0.683762452
#> disp          0.002947114   0.002947865   0.002948572   0.002949239
#> drat          7.597961129   7.597293648   7.596645097   7.596014822
#> gear          8.999178203   8.998390948   8.997626737   8.996884832
#> hp            0.032783314   0.032778789   0.032774352   0.032770003
#> qsec          2.439987267   2.439844355   2.439703551   2.439564906
#> vs           -7.616662968  -7.615985416  -7.615319014  -7.614663570
#> wt           -1.996727609  -1.996889799  -1.997044005  -1.997190736
#>                    0.0123        0.0114        0.0107        0.0100
#> (Intercept) -72.514419067 -72.507097141 -72.499973471 -72.493042661
#> am           -3.352001676  -3.351570889  -3.351150520  -3.350740220
#> carb         -3.220449420  -3.220170033  -3.219899011  -3.219636043
#> cyl           0.683658532   0.683560351   0.683467664   0.683380230
#> disp          0.002949868   0.002950464   0.002951028   0.002951563
#> drat          7.595402188   7.594806583   7.594227415   7.593664111
#> gear          8.996164510   8.995465068   8.994785823   8.994126108
#> hp            0.032765739   0.032761559   0.032757460   0.032753443
#> qsec          2.439428461   2.439294254   2.439162313   2.439032660
#> vs           -7.614018901  -7.613384829  -7.612761180  -7.612147786
#> wt           -1.997330468  -1.997463653  -1.997590715  -1.997712052
#>                    0.0093        0.0087        0.0081        0.0075
#> (Intercept) -72.486299389 -72.479738414 -72.473354579 -72.467142820
#> am           -3.350339654  -3.349948501  -3.349566451  -3.349193209
#> carb         -3.219380831  -3.219133085  -3.218892527  -3.218658889
#> cyl           0.683297814   0.683220191   0.683147139   0.683078445
#> disp          0.002952071   0.002952555   0.002953017   0.002953459
#> drat          7.593116120   7.592582910   7.592063966   7.591558793
#> gear          8.993485274   8.992862692   8.992257750   8.991669856
#> hp            0.032749503   0.032745642   0.032741855   0.032738143
#> qsec          2.438905311   2.438780278   2.438657565   2.438537174
#> vs           -7.611544485  -7.610951116  -7.610367527  -7.609793565
#> wt           -1.997828041  -1.997939033  -1.998045359  -1.998147331
#>                    0.0070        0.0066        0.0061        0.0057
#> (Intercept) -72.461098166 -72.455215744 -72.449490785 -72.443918621
#> am           -3.348828491  -3.348472024  -3.348123547  -3.347782810
#> carb         -3.218431913  -3.218211351  -3.217996964  -3.217788523
#> cyl           0.683013903   0.682953312   0.682896479   0.682843217
#> disp          0.002953882   0.002954289   0.002954679   0.002955056
#> drat          7.591066914   7.590587870   7.590121217   7.589666531
#> gear          8.991098432   8.990542923   8.990002787   8.989477503
#> hp            0.032734504   0.032730936   0.032727437   0.032724007
#> qsec          2.438419102   2.438303340   2.438189879   2.438078703
#> vs           -7.609229083  -7.608673938  -7.608127988  -7.607591096
#> wt           -1.998245238  -1.998339354  -1.998429932  -1.998517213
#>                    0.0053
#> (Intercept) -72.438494693
#> am           -3.347449572
#> carb         -3.217585808
#> cyl           0.682793347
#> disp          0.002955421
#> drat          7.589223402
#> gear          8.988966565
#> hp            0.032720644
#> qsec          2.437969796
#> vs           -7.607063127
#> wt           -1.998601418
#> 
#> $iter
#>   [1]   0   8   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1   1   4   5   5   5   5   5   5 180  31   1   1   1
#>  [37]   1  70   1   1   1   1   1   1  35   1   1   1   1   1   1 121   1   1
#>  [55]  72   1   1 318   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1 101   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.313814072 4.955680788 4.621684489 4.310198423 4.019705475 3.748790779
#>   [7] 3.496134827 3.260507041 3.040759779 2.835822747 2.644697785 2.466454006
#>  [13] 2.300223261 2.145195912 2.000616887 1.865782005 1.740034542 1.622762038
#>  [19] 1.513393307 1.411395664 1.316272320 1.227559971 1.144826537 1.067669060
#>  [25] 0.995711739 0.928604100 0.866019292 0.807652490 0.753219415 0.702454948
#>  [31] 0.655111835 0.610959490 0.569782865 0.531381407 0.495568079 0.462168449
#>  [37] 0.431019842 0.401970547 0.374879078 0.349613483 0.326050704 0.304075978
#>  [43] 0.283582275 0.264469779 0.246645401 0.230022326 0.214519591 0.200061689
#>  [49] 0.186578200 0.174003454 0.162276204 0.151339331 0.141139566 0.131627232
#>  [55] 0.122755997 0.114482654 0.106766906 0.099571174 0.092860410 0.086601929
#>  [61] 0.080765249 0.075321942 0.070245495 0.065511184 0.061095949 0.056978287
#>  [67] 0.053138141 0.049556808 0.046216845 0.043101984 0.040197055 0.037487908
#>  [73] 0.034961348 0.032605070 0.030407598 0.028358227 0.026446978 0.024664540
#>  [79] 0.023002233 0.021451959 0.020006169 0.018657820 0.017400345 0.016227620
#>  [85] 0.015133933 0.014113957 0.013162723 0.012275600 0.011448265 0.010676691
#>  [91] 0.009957117 0.009286041 0.008660193 0.008076525 0.007532194 0.007024549
#>  [97] 0.006551118 0.006109595 0.005697829 0.005313814
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 31
#> 
#> $loss
#>   [1] 773.99143 660.15931 564.92181 485.60303 419.89270 365.79746 321.59843
#>   [8] 285.81454 257.17055 234.56940 217.06819 203.85734 194.24249 187.62888
#>  [15] 183.50773 181.44446 181.02241 181.02241 181.02241 181.02241 181.02241
#>  [22] 181.02241 181.02241 181.02241 180.60713 175.18312 170.71492 167.06169
#>  [29] 164.10159 161.72969 159.85573 141.00875 127.85074 127.85074 127.85073
#>  [36] 127.85073 127.85073  91.48128  91.48128  91.48127  91.48127  91.48127
#>  [43]  91.48127  91.48127  84.87374  84.87373  84.87372  84.87372  84.87371
#>  [50]  84.87371  84.87371  74.47851  74.47850  74.47849  71.59214  71.59213
#>  [57]  71.59212  64.34040  64.34037  64.34034  64.34031  64.34029  64.34026
#>  [64]  64.34024  64.34022  64.34020  64.34017  64.34015  64.34013  64.34012
#>  [71]  64.34010  64.34008  64.34006  64.34005  64.34003  64.34002  64.34000
#>  [78]  64.33999  64.33998  64.33996  64.33995  64.33994  64.19232  64.19230
#>  [85]  64.19228  64.19225  64.19223  64.19221  64.19220  64.19218  64.19216
#>  [92]  64.19215  64.19213  64.19212  64.19210  64.19209  64.19208  64.19207
#>  [99]  64.19206  64.19205
#> 
#> $linear.predictors
#>      5.3138   4.9557   4.6217   4.3102   4.0197   3.7488   3.4961   3.2605
#> 1  19.24286 19.58909 19.91199 20.21313 20.49398 20.75589 21.00015 21.22795
#> 2  19.24286 19.80735 20.33380 20.82477 21.28265 21.70967 22.10791 22.47931
#> 3  19.24286 18.74964 18.28966 17.86068 17.46061 17.08751 16.73955 16.41504
#> 4  19.24286 18.74964 18.28966 17.86068 17.46061 17.08751 16.73955 16.41504
#> 5  19.24286 19.64492 20.01988 20.36957 20.69569 20.99984 21.28348 21.54801
#> 6  19.24286 19.55719 19.85035 20.12374 20.37871 20.61649 20.83825 21.04506
#> 7  19.24286 19.10305 18.97266 18.85106 18.73766 18.63190 18.53326 18.44128
#> 8  19.24286 19.10305 18.97266 18.85106 18.73766 18.63190 18.53326 18.44128
#> 9  19.24286 19.10305 18.97266 18.85106 18.73766 18.63190 18.53326 18.44128
#> 10 19.24286 18.27954 17.38115 16.54330 15.76193 15.03321 14.35361 13.71981
#> 11 19.24286 18.32991 17.47849 16.68445 15.94393 15.25332 14.60925 14.00859
#> 12 19.24286 18.41385 17.64072 16.91970 16.24727 15.62015 15.03531 14.48988
#> 13 19.24286 19.93033 20.57147 21.16941 21.72704 22.24709 22.73209 23.18440
#> 14 19.24286 19.94292 20.59581 21.20469 21.77254 22.30211 22.79600 23.25660
#> 15 19.24286 19.75656 20.23565 20.68245 21.09913 21.48773 21.85014 22.18813
#> 16 19.24286 18.98468 18.74391 18.51937 18.30995 18.11465 17.93252 17.76266
#> 17 19.24286 18.79161 18.37077 17.97830 17.61228 17.27093 16.95258 16.65569
#> 18 19.24286 19.86150 20.43844 20.97650 21.47830 21.94628 22.38272 22.78974
#> 19 19.24286 19.65205 20.03367 20.38957 20.72148 21.03102 21.31970 21.58892
#> 20 19.24286 18.99728 18.76825 18.55465 18.35545 18.16968 17.99643 17.83485
#> 21 19.24286 19.75279 20.22835 20.67186 21.08548 21.47123 21.83097 22.16647
#>      3.0408   2.8358   2.6447   2.4665   2.3002   2.1452   2.0006    1.8658
#> 1  21.44040 21.63853 21.82331 21.99563 22.15634 22.30621 22.44599 22.576346
#> 2  22.82568 23.14870 23.44996 23.73091 23.99292 24.23728 24.46517 24.677696
#> 3  16.11241 15.83017 15.56695 15.32147 15.09254 14.87904 14.67992 14.494228
#> 4  16.11241 15.83017 15.56695 15.32147 15.09254 14.87904 14.67992 14.494228
#> 5  21.79471 22.02479 22.23935 22.43946 22.62608 22.80012 22.96243 23.113807
#> 6  21.23794 21.41781 21.58556 21.74201 21.88791 22.02398 22.15088 22.269225
#> 7  18.35549 18.27549 18.20088 18.13129 18.06640 18.00588 17.94944 17.896800
#> 8  18.35549 18.27549 18.20088 18.13129 18.06640 18.00588 17.94944 17.896800
#> 9  18.35549 18.27549 18.20088 18.13129 18.06640 18.00588 17.94944 17.896800
#> 10 13.12873 12.57749 12.06339 11.58395 11.13681 10.71982 10.33092  9.968242
#> 11 13.44841 12.92599 12.43877 11.98440 11.56064 11.16545 10.79689 10.453169
#> 12 13.98121 13.50682 13.06441 12.65181 12.26702 11.90817 11.57350 11.261381
#> 13 23.60623 23.99963 24.36651 24.70867 25.02777 25.32536 25.60290 25.861727
#> 14 23.68615 24.08675 24.46036 24.80878 25.13373 25.43677 25.71939 25.982959
#> 15 22.50334 22.79730 23.07145 23.32712 23.56556 23.78794 23.99532 24.188728
#> 16 17.60425 17.45651 17.31873 17.19024 17.07040 16.95865 16.85442 16.757221
#> 17 16.37881 16.12059 15.87977 15.65518 15.44573 15.25040 15.06823 14.898334
#> 18 23.16933 23.52334 23.85349 24.16139 24.44854 24.71633 24.96608 25.198993
#> 19 21.84000 22.07416 22.29253 22.49619 22.68612 22.86325 23.02845 23.182505
#> 20 17.68417 17.54363 17.41258 17.29035 17.17636 17.07005 16.97091 16.878453
#> 21 22.47936 22.77116 23.04329 23.29709 23.53378 23.75451 23.96037 24.152359
#>       1.7400    1.6228    1.5134    1.4114    1.3163    1.2276    1.1448
#> 1  22.667717 22.667717 22.667717 22.667717 22.667717 22.667717 22.667717
#> 2  24.826666 24.826666 24.826666 24.826666 24.826666 24.826666 24.826666
#> 3  14.364068 14.364068 14.364068 14.364068 14.364068 14.364068 14.364068
#> 4  14.364068 14.364068 14.364068 14.364068 14.364068 14.364068 14.364068
#> 5  23.219910 23.219910 23.219910 23.219910 23.219910 23.219910 23.219910
#> 6  22.352178 22.352178 22.352178 22.352178 22.352178 22.352178 22.352178
#> 7  17.859904 17.859904 17.859904 17.859904 17.859904 17.859904 17.859904
#> 8  17.859904 17.859904 17.859904 17.859904 17.859904 17.859904 17.859904
#> 9  17.859904 17.859904 17.859904 17.859904 17.859904 17.859904 17.859904
#> 10  9.714025  9.714025  9.714025  9.714025  9.714025  9.714025  9.714025
#> 11 10.212244 10.212244 10.212244 10.212244 10.212244 10.212244 10.212244
#> 12 11.042609 11.042609 11.042609 11.042609 11.042609 11.042609 11.042609
#> 13 26.043150 26.043150 26.043150 26.043150 26.043150 26.043150 26.043150
#> 14 26.167705 26.167705 26.167705 26.167705 26.167705 26.167705 26.167705
#> 15 24.324295 24.324295 24.324295 24.324295 24.324295 24.324295 24.324295
#> 16 16.689090 16.689090 16.689090 16.689090 16.689090 16.689090 16.689090
#> 17 14.779250 14.779250 14.779250 14.779250 14.779250 14.779250 14.779250
#> 18 25.362251 25.362251 25.362251 25.362251 25.362251 25.362251 25.362251
#> 19 23.290491 23.290491 23.290491 23.290491 23.290491 23.290491 23.290491
#> 20 16.813644 16.813644 16.813644 16.813644 16.813644 16.813644 16.813644
#> 21 24.286928 24.286928 24.286928 24.286928 24.286928 24.286928 24.286928
#>       1.0677    0.9957    0.9286   0.8660   0.8077    0.7532    0.7025
#> 1  22.667717 22.659909 22.551330 22.45006 22.35561 22.267529 22.185384
#> 2  24.826666 24.834969 24.950444 25.05815 25.15859 25.252267 25.339629
#> 3  14.364068 14.374726 14.522942 14.66118 14.79011 14.910343 15.022475
#> 4  14.364068 14.362844 14.345821 14.32994 14.31514 14.301327 14.288448
#> 5  23.219910 23.223546 23.274117 23.32129 23.36527 23.406298 23.444558
#> 6  22.352178 22.344621 22.239520 22.14149 22.05007 21.964813 21.885300
#> 7  17.859904 17.861849 17.888901 17.91413 17.93766 17.959606 17.980072
#> 8  17.859904 17.861849 17.888901 17.91413 17.93766 17.959606 17.980072
#> 9  17.859904 17.861849 17.888901 17.91413 17.93766 17.959606 17.980072
#> 10  9.714025  9.716487  9.750736  9.78268  9.81247  9.840253  9.866164
#> 11 10.212244 10.214311 10.243066 10.26989 10.29490 10.318225 10.339980
#> 12 11.042609 11.044018 11.063617 11.08190 11.09895 11.114846 11.129673
#> 13 26.043150 26.050489 26.152551 26.24774 26.33652 26.419316 26.496530
#> 14 26.167705 26.169004 26.187073 26.20393 26.21964 26.234300 26.247970
#> 15 24.324295 24.332996 24.454011 24.56688 24.67214 24.770312 24.861864
#> 16 16.689090 16.697904 16.820485 16.93482 17.04144 17.140880 17.233618
#> 17 14.779250 14.777697 14.756096 14.73595 14.71716 14.699637 14.683295
#> 18 25.362251 25.364189 25.391139 25.41627 25.43972 25.461579 25.481967
#> 19 23.290491 23.270307 22.989621 22.72783 22.48367 22.255977 22.043627
#> 20 16.813644 16.786715 16.412203 16.06290 15.73713 15.433322 15.149989
#> 21 24.286928 24.289719 24.328525 24.36472 24.39848 24.429955 24.459314
#>       0.6551   0.6110   0.5698   0.5314   0.4956   0.4622   0.4310   0.4020
#> 1  22.108775 22.22576 22.63463 22.63474 22.63481 22.63485 22.63488 20.38682
#> 2  25.421102 26.17249 27.44001 27.44039 27.44063 27.44079 27.44089 27.02442
#> 3  15.127050 15.54286 16.10336 16.10353 16.10365 16.10372 16.10377 17.35555
#> 4  14.276437 14.10914 13.81241 13.81232 13.81226 13.81223 13.81220 13.02811
#> 5  23.480239 22.91055 21.71870 21.71836 21.71814 21.71800 21.71791 24.16508
#> 6  21.811146 20.75492 18.87754 18.87698 18.87663 18.87640 18.87625 20.46004
#> 7  17.999158 17.73439 17.17453 17.17438 17.17427 17.17421 17.17416 14.78911
#> 8  17.999158 17.73439 17.17453 17.17438 17.17427 17.17421 17.17416 14.78911
#> 9  17.999158 17.73439 17.17453 17.17438 17.17427 17.17421 17.17416 14.78911
#> 10  9.890328 10.24049 10.86389 10.86407 10.86419 10.86427 10.86432 12.00300
#> 11 10.360268 10.65499 11.17980 11.17996 11.18006 11.18012 11.18016 12.25928
#> 12 11.143502 11.34582 11.70632 11.70643 11.70650 11.70654 11.70657 13.10134
#> 13 26.568540 27.18456 28.21137 28.21167 28.21187 28.21200 28.21208 27.86647
#> 14 26.260719 26.57132 27.14487 27.14503 27.14514 27.14521 27.14526 28.70488
#> 15 24.947246 24.54622 23.56445 23.56418 23.56400 23.56389 23.56381 21.64272
#> 16 17.320105 17.47718 17.57762 17.57766 17.57768 17.57770 17.57771 17.35555
#> 17 14.668054 14.45455 14.07567 14.07556 14.07548 14.07544 14.07541 14.93190
#> 18 25.500982 25.90122 26.63414 26.63436 26.63450 26.63459 26.63464 29.29050
#> 19 21.845588 21.31017 20.73857 20.73838 20.73826 20.73818 20.73812 19.64713
#> 20 14.885751 14.48798 14.34076 14.34068 14.34063 14.34060 14.34058 14.80714
#> 21 24.486694 25.00659 25.95230 25.95257 25.95275 25.95287 25.95295 25.70276
#>      0.3749   0.3496   0.3261   0.3041   0.2836   0.2645   0.2466   0.2300
#> 1  20.38667 20.38659 20.38654 20.38651 20.38650 20.38649 20.63342 20.63359
#> 2  27.02460 27.02470 27.02476 27.02479 27.02481 27.02482 26.83551 26.83518
#> 3  17.35570 17.35579 17.35583 17.35586 17.35588 17.35588 17.75951 17.75960
#> 4  13.02803 13.02799 13.02796 13.02795 13.02794 13.02794 13.97910 13.97963
#> 5  24.16523 24.16531 24.16536 24.16539 24.16540 24.16541 23.46879 23.46834
#> 6  20.45988 20.45980 20.45975 20.45972 20.45970 20.45970 20.03646 20.03639
#> 7  14.78919 14.78924 14.78926 14.78928 14.78928 14.78929 14.89991 14.89997
#> 8  14.78919 14.78924 14.78926 14.78928 14.78928 14.78929 15.29362 15.29383
#> 9  14.78919 14.78924 14.78926 14.78928 14.78928 14.78929 15.23572 15.23591
#> 10 12.00304 12.00306 12.00307 12.00308 12.00308 12.00308 11.23152 11.23132
#> 11 12.25929 12.25929 12.25929 12.25930 12.25930 12.25930 11.23058 11.23030
#> 12 13.10125 13.10120 13.10117 13.10115 13.10114 13.10114 11.98099 11.98074
#> 13 27.86656 27.86661 27.86663 27.86665 27.86666 27.86666 27.63340 27.63312
#> 14 28.70450 28.70429 28.70417 28.70410 28.70406 28.70404 28.84511 28.84531
#> 15 21.64274 21.64275 21.64276 21.64276 21.64277 21.64277 22.36513 22.36538
#> 16 17.35570 17.35579 17.35583 17.35586 17.35588 17.35588 17.76530 17.76539
#> 17 14.93160 14.93143 14.93133 14.93128 14.93125 14.93123 15.15621 15.15660
#> 18 29.29084 29.29103 29.29114 29.29120 29.29123 29.29125 29.51265 29.51229
#> 19 19.64712 19.64711 19.64711 19.64711 19.64711 19.64711 20.02379 20.02387
#> 20 14.80695 14.80685 14.80679 14.80676 14.80674 14.80673 15.06645 15.06668
#> 21 25.70273 25.70271 25.70270 25.70269 25.70269 25.70269 25.14683 25.14655
#>      0.2145   0.2001   0.1866   0.1740   0.1623   0.1513   0.1411   0.1316
#> 1  20.63373 20.63383 20.63391 20.63398 20.63403 20.63084 20.63077 20.63071
#> 2  26.83493 26.83473 26.83457 26.83445 26.83436 26.20677 26.20678 26.20678
#> 3  17.75967 17.75973 17.75977 17.75980 17.75983 18.12313 18.12336 18.12357
#> 4  13.98003 13.98036 13.98061 13.98081 13.98096 12.89868 12.89843 12.89820
#> 5  23.46799 23.46771 23.46750 23.46733 23.46719 23.64415 23.64411 23.64407
#> 6  20.03633 20.03628 20.03625 20.03622 20.03620 21.13220 21.13229 21.13238
#> 7  14.90002 14.90006 14.90009 14.90012 14.90014 15.15709 15.15718 15.15725
#> 8  15.29400 15.29413 15.29423 15.29431 15.29437 16.00859 16.00868 16.00877
#> 9  15.23606 15.23618 15.23627 15.23634 15.23640 16.37319 16.37333 16.37345
#> 10 11.23117 11.23105 11.23095 11.23088 11.23082 11.17932 11.17929 11.17926
#> 11 11.23009 11.22992 11.22979 11.22968 11.22960 10.94787 10.94785 10.94782
#> 12 11.98055 11.98040 11.98028 11.98018 11.98011 11.52381 11.52380 11.52379
#> 13 27.63290 27.63273 27.63259 27.63249 27.63241 28.29515 28.29530 28.29543
#> 14 28.84546 28.84558 28.84568 28.84575 28.84581 29.62769 29.62774 29.62780
#> 15 22.36558 22.36573 22.36585 22.36595 22.36602 21.63958 21.63917 21.63879
#> 16 17.76547 17.76552 17.76557 17.76560 17.76563 18.45169 18.45195 18.45218
#> 17 15.15690 15.15714 15.15733 15.15748 15.15760 13.91941 13.91926 13.91912
#> 18 29.51200 29.51178 29.51161 29.51147 29.51137 28.81090 28.81102 28.81112
#> 19 20.02394 20.02399 20.02403 20.02406 20.02408 19.45721 19.45706 19.45693
#> 20 15.06686 15.06700 15.06711 15.06719 15.06726 15.65340 15.65341 15.65343
#> 21 25.14632 25.14615 25.14601 25.14590 25.14582 24.41932 24.41922 24.41913
#>      0.1228   0.1145   0.1068   0.0996   0.0929   0.0866   0.0808   0.0753
#> 1  21.61275 21.61252 21.61231 20.82046 20.82035 20.82025 20.82015 20.82005
#> 2  26.10575 26.10577 26.10579 25.44332 25.44321 25.44310 25.44300 25.44289
#> 3  18.17519 18.17539 18.17558 18.74281 18.74296 18.74311 18.74326 18.74340
#> 4  12.51111 12.51102 12.51095 11.96613 11.96590 11.96568 11.96546 11.96524
#> 5  23.71183 23.71175 23.71167 25.03364 25.03391 25.03417 25.03443 25.03468
#> 6  20.16482 20.16504 20.16523 19.19704 19.19687 19.19671 19.19654 19.19638
#> 7  15.18925 15.18933 15.18940 14.80535 14.80529 14.80523 14.80518 14.80513
#> 8  16.04600 16.04609 16.04617 15.86873 15.86869 15.86865 15.86861 15.86857
#> 9  16.49522 16.49533 16.49543 16.73354 16.73359 16.73363 16.73367 16.73371
#> 10 11.25972 11.25967 11.25962 10.62377 10.62366 10.62355 10.62345 10.62334
#> 11 11.04616 11.04610 11.04605 10.82430 10.82428 10.82426 10.82424 10.82423
#> 12 11.63982 11.63977 11.63972 12.29021 12.29037 12.29052 12.29067 12.29082
#> 13 28.45825 28.45832 28.45838 28.52579 28.52586 28.52593 28.52600 28.52606
#> 14 29.80875 29.80873 29.80872 29.99606 29.99614 29.99622 29.99630 29.99638
#> 15 21.45552 21.45523 21.45497 21.89604 21.89595 21.89586 21.89578 21.89570
#> 16 18.55892 18.55914 18.55934 18.55378 18.55384 18.55391 18.55397 18.55402
#> 17 13.71177 13.71170 13.71164 14.46151 14.46160 14.46170 14.46179 14.46188
#> 18 28.43597 28.43614 28.43630 28.54991 28.54996 28.55001 28.55006 28.55011
#> 19 20.05991 20.05969 20.05949 19.87539 19.87535 19.87531 19.87527 19.87523
#> 20 15.29407 15.29414 15.29421 16.03460 16.03468 16.03476 16.03483 16.03491
#> 21 24.35922 24.35912 24.35904 23.85763 23.85754 23.85745 23.85736 23.85727
#>      0.0702   0.0655   0.0611   0.0570   0.0531   0.0496   0.0462   0.0431
#> 1  20.81995 20.81986 20.81977 20.81967 20.81958 20.81950 20.81941 20.81933
#> 2  25.44279 25.44269 25.44259 25.44250 25.44240 25.44231 25.44222 25.44213
#> 3  18.74354 18.74368 18.74382 18.74395 18.74408 18.74421 18.74434 18.74446
#> 4  11.96503 11.96482 11.96462 11.96442 11.96422 11.96403 11.96384 11.96366
#> 5  25.03492 25.03516 25.03540 25.03563 25.03586 25.03608 25.03630 25.03651
#> 6  19.19623 19.19607 19.19592 19.19577 19.19563 19.19549 19.19535 19.19521
#> 7  14.80507 14.80502 14.80497 14.80492 14.80487 14.80483 14.80478 14.80474
#> 8  15.86853 15.86849 15.86846 15.86842 15.86838 15.86835 15.86831 15.86828
#> 9  16.73375 16.73379 16.73383 16.73387 16.73391 16.73394 16.73398 16.73401
#> 10 10.62324 10.62314 10.62304 10.62295 10.62285 10.62276 10.62267 10.62258
#> 11 10.82421 10.82419 10.82418 10.82416 10.82415 10.82413 10.82412 10.82410
#> 12 12.29097 12.29111 12.29124 12.29138 12.29151 12.29164 12.29177 12.29190
#> 13 28.52613 28.52619 28.52625 28.52632 28.52637 28.52643 28.52649 28.52655
#> 14 29.99645 29.99653 29.99660 29.99667 29.99674 29.99681 29.99688 29.99694
#> 15 21.89562 21.89554 21.89546 21.89538 21.89531 21.89524 21.89517 21.89510
#> 16 18.55408 18.55414 18.55419 18.55425 18.55430 18.55435 18.55440 18.55445
#> 17 14.46196 14.46205 14.46213 14.46221 14.46229 14.46237 14.46245 14.46252
#> 18 28.55015 28.55020 28.55025 28.55029 28.55033 28.55038 28.55042 28.55046
#> 19 19.87519 19.87516 19.87512 19.87509 19.87505 19.87502 19.87498 19.87495
#> 20 16.03499 16.03506 16.03513 16.03520 16.03527 16.03534 16.03541 16.03547
#> 21 23.85719 23.85710 23.85702 23.85694 23.85686 23.85679 23.85671 23.85664
#>      0.0402   0.0375   0.0350   0.0326   0.0304   0.0284   0.0264   0.0247
#> 1  20.81924 20.81916 20.81908 20.81901 20.81893 20.81886 20.81878 20.81871
#> 2  25.44205 25.44196 25.44188 25.44180 25.44172 25.44164 25.44157 25.44149
#> 3  18.74458 18.74470 18.74481 18.74493 18.74504 18.74515 18.74525 18.74536
#> 4  11.96348 11.96330 11.96313 11.96296 11.96279 11.96263 11.96247 11.96231
#> 5  25.03672 25.03693 25.03713 25.03733 25.03752 25.03771 25.03790 25.03808
#> 6  19.19508 19.19495 19.19482 19.19470 19.19457 19.19445 19.19433 19.19422
#> 7  14.80469 14.80465 14.80460 14.80456 14.80452 14.80448 14.80444 14.80440
#> 8  15.86825 15.86822 15.86818 15.86815 15.86812 15.86809 15.86806 15.86804
#> 9  16.73405 16.73408 16.73411 16.73415 16.73418 16.73421 16.73424 16.73427
#> 10 10.62250 10.62241 10.62233 10.62225 10.62217 10.62209 10.62201 10.62194
#> 11 10.82409 10.82408 10.82406 10.82405 10.82404 10.82402 10.82401 10.82400
#> 12 12.29202 12.29214 12.29226 12.29237 12.29248 12.29259 12.29270 12.29281
#> 13 28.52660 28.52666 28.52671 28.52676 28.52681 28.52686 28.52691 28.52696
#> 14 29.99701 29.99707 29.99713 29.99719 29.99725 29.99731 29.99737 29.99742
#> 15 21.89503 21.89496 21.89490 21.89483 21.89477 21.89471 21.89465 21.89459
#> 16 18.55450 18.55455 18.55460 18.55464 18.55469 18.55473 18.55478 18.55482
#> 17 14.46260 14.46267 14.46274 14.46281 14.46288 14.46295 14.46301 14.46308
#> 18 28.55050 28.55054 28.55058 28.55062 28.55065 28.55069 28.55073 28.55076
#> 19 19.87492 19.87489 19.87486 19.87483 19.87480 19.87477 19.87474 19.87471
#> 20 16.03553 16.03560 16.03566 16.03572 16.03578 16.03583 16.03589 16.03595
#> 21 23.85657 23.85649 23.85643 23.85636 23.85629 23.85623 23.85616 23.85610
#>      0.0230   0.0215   0.0200   0.0187   0.0174   0.0162   0.0151   0.0141
#> 1  20.81864 20.81857 20.81851 20.81844 20.81271 20.81286 20.81300 20.81315
#> 2  25.44142 25.44135 25.44127 25.44121 25.39019 25.39024 25.39030 25.39035
#> 3  18.74546 18.74556 18.74566 18.74575 18.87293 18.87277 18.87261 18.87246
#> 4  11.96216 11.96201 11.96186 11.96171 12.04113 12.04141 12.04167 12.04193
#> 5  25.03826 25.03843 25.03860 25.03877 25.05912 25.05895 25.05878 25.05862
#> 6  19.19411 19.19399 19.19389 19.19378 19.14560 19.14560 19.14561 19.14562
#> 7  14.80437 14.80433 14.80429 14.80426 14.62833 14.62830 14.62827 14.62825
#> 8  15.86801 15.86798 15.86795 15.86793 15.79519 15.79519 15.79518 15.79518
#> 9  16.73430 16.73433 16.73436 16.73438 16.67141 16.67134 16.67128 16.67121
#> 10 10.62186 10.62179 10.62172 10.62165 10.79964 10.79979 10.79993 10.80007
#> 11 10.82399 10.82397 10.82396 10.82395 10.88625 10.88629 10.88632 10.88636
#> 12 12.29291 12.29302 12.29312 12.29321 12.24851 12.24837 12.24824 12.24811
#> 13 28.52700 28.52705 28.52710 28.52714 28.50438 28.50427 28.50417 28.50407
#> 14 29.99748 29.99753 29.99758 29.99764 30.12323 30.12317 30.12311 30.12305
#> 15 21.89453 21.89447 21.89441 21.89436 21.89821 21.89846 21.89869 21.89891
#> 16 18.55486 18.55490 18.55494 18.55498 18.58146 18.58133 18.58120 18.58109
#> 17 14.46314 14.46320 14.46326 14.46332 14.37463 14.37456 14.37449 14.37443
#> 18 28.55079 28.55083 28.55086 28.55089 28.61665 28.61656 28.61647 28.61638
#> 19 19.87468 19.87466 19.87463 19.87461 19.82877 19.82888 19.82898 19.82909
#> 20 16.03600 16.03605 16.03610 16.03616 16.05772 16.05766 16.05760 16.05754
#> 21 23.85604 23.85598 23.85592 23.85586 23.76394 23.76401 23.76407 23.76413
#>      0.0132   0.0123   0.0114   0.0107   0.0100   0.0093   0.0087   0.0081
#> 1  20.81328 20.81342 20.81355 20.81367 20.81379 20.81391 20.81402 20.81413
#> 2  25.39041 25.39046 25.39051 25.39056 25.39061 25.39066 25.39071 25.39076
#> 3  18.87232 18.87217 18.87203 18.87190 18.87177 18.87164 18.87152 18.87140
#> 4  12.04218 12.04242 12.04265 12.04288 12.04310 12.04331 12.04351 12.04371
#> 5  25.05845 25.05830 25.05814 25.05799 25.05784 25.05770 25.05756 25.05742
#> 6  19.14564 19.14566 19.14568 19.14571 19.14574 19.14577 19.14581 19.14584
#> 7  14.62823 14.62821 14.62819 14.62817 14.62816 14.62815 14.62814 14.62813
#> 8  15.79518 15.79518 15.79518 15.79519 15.79519 15.79519 15.79519 15.79519
#> 9  16.67115 16.67109 16.67103 16.67097 16.67092 16.67086 16.67081 16.67076
#> 10 10.80020 10.80033 10.80045 10.80057 10.80068 10.80079 10.80090 10.80100
#> 11 10.88639 10.88642 10.88645 10.88648 10.88651 10.88653 10.88656 10.88658
#> 12 12.24799 12.24787 12.24775 12.24764 12.24753 12.24742 12.24731 12.24721
#> 13 28.50397 28.50388 28.50379 28.50371 28.50362 28.50354 28.50346 28.50339
#> 14 30.12299 30.12294 30.12288 30.12283 30.12278 30.12273 30.12268 30.12264
#> 15 21.89913 21.89933 21.89952 21.89970 21.89988 21.90005 21.90021 21.90036
#> 16 18.58097 18.58086 18.58076 18.58066 18.58056 18.58047 18.58038 18.58030
#> 17 14.37436 14.37430 14.37424 14.37418 14.37412 14.37406 14.37401 14.37395
#> 18 28.61630 28.61623 28.61615 28.61608 28.61601 28.61594 28.61588 28.61582
#> 19 19.82919 19.82928 19.82938 19.82946 19.82955 19.82963 19.82971 19.82979
#> 20 16.05748 16.05742 16.05736 16.05730 16.05724 16.05718 16.05713 16.05707
#> 21 23.76419 23.76425 23.76430 23.76436 23.76441 23.76447 23.76452 23.76457
#>      0.0075   0.0070   0.0066   0.0061   0.0057   0.0053
#> 1  20.81424 20.81434 20.81444 20.81453 20.81463 20.81472
#> 2  25.39080 25.39085 25.39089 25.39094 25.39098 25.39102
#> 3  18.87128 18.87117 18.87106 18.87096 18.87085 18.87075
#> 4  12.04390 12.04408 12.04426 12.04443 12.04460 12.04476
#> 5  25.05728 25.05715 25.05702 25.05689 25.05676 25.05664
#> 6  19.14588 19.14593 19.14597 19.14601 19.14606 19.14611
#> 7  14.62812 14.62811 14.62811 14.62810 14.62810 14.62810
#> 8  15.79520 15.79520 15.79520 15.79521 15.79521 15.79522
#> 9  16.67071 16.67066 16.67061 16.67056 16.67052 16.67047
#> 10 10.80110 10.80120 10.80129 10.80139 10.80148 10.80156
#> 11 10.88661 10.88663 10.88665 10.88668 10.88670 10.88672
#> 12 12.24711 12.24701 12.24692 12.24682 12.24673 12.24664
#> 13 28.50331 28.50324 28.50317 28.50311 28.50304 28.50298
#> 14 30.12259 30.12255 30.12251 30.12247 30.12243 30.12239
#> 15 21.90050 21.90064 21.90077 21.90090 21.90101 21.90113
#> 16 18.58022 18.58014 18.58006 18.57999 18.57993 18.57986
#> 17 14.37390 14.37385 14.37380 14.37375 14.37370 14.37365
#> 18 28.61576 28.61570 28.61565 28.61560 28.61555 28.61550
#> 19 19.82986 19.82993 19.82999 19.83006 19.83012 19.83018
#> 20 16.05701 16.05696 16.05691 16.05685 16.05680 16.05675
#> 21 23.76462 23.76467 23.76472 23.76476 23.76481 23.76485
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 22.8 18.7 14.3 24.4 19.2 16.4 17.3 15.2 10.4 10.4 14.7 32.4 30.4 21.5
#> [16] 15.2 13.3 30.4 19.7 15.0 21.4
#> 
#> $X
#>               am        carb        cyl       disp        drat       gear
#>  [1,]  1.2747549  0.51323935 -0.2102353 -0.6445201  0.70846655  0.5275044
#>  [2,]  1.2747549 -1.28309838 -1.3139708 -1.0508099  0.60486071  0.5275044
#>  [3,] -0.7844645 -0.68431914  0.8935002  0.9181332 -0.84562093 -0.8571946
#>  [4,] -0.7844645  0.51323935  0.8935002  0.9181332 -0.72129394 -0.8571946
#>  [5,] -0.7844645 -0.68431914 -1.3139708 -0.7484365  0.27332205  0.5275044
#>  [6,] -0.7844645  0.51323935 -0.2102353 -0.5851392  0.74990888  0.5275044
#>  [7,] -0.7844645 -0.08553989  0.8935002  0.2602562 -1.01139027 -0.8571946
#>  [8,] -0.7844645 -0.08553989  0.8935002  0.2602562 -1.01139027 -0.8571946
#>  [9,] -0.7844645 -0.08553989  0.8935002  0.2602562 -1.01139027 -0.8571946
#> [10,] -0.7844645  0.51323935  0.8935002  1.7932190 -1.30148660 -0.8571946
#> [11,] -0.7844645  0.51323935  0.8935002  1.6994598 -1.15643843 -0.8571946
#> [12,] -0.7844645  0.51323935  0.8935002  1.5431945 -0.67985160 -0.8571946
#> [13,]  1.2747549 -1.28309838 -1.3139708 -1.2797386  1.08144754  0.5275044
#> [14,]  1.2747549 -0.68431914 -1.3139708 -1.3031784  2.84274669  0.5275044
#> [15,] -0.7844645 -1.28309838 -1.3139708 -0.9562694  0.29404322 -0.8571946
#> [16,] -0.7844645 -0.68431914  0.8935002  0.4805903 -0.84562093 -0.8571946
#> [17,] -0.7844645  0.51323935  0.8935002  0.8400005  0.35620672 -0.8571946
#> [18,]  1.2747549 -0.68431914 -1.3139708 -1.1516010  0.43909138  1.9122034
#> [19,]  1.2747549  1.71079785 -0.2102353 -0.7617190  0.12827389  1.9122034
#> [20,]  1.2747549  2.90835634  0.8935002  0.4571505 -0.03749544  1.9122034
#> [21,]  1.2747549 -0.68431914 -1.3139708 -0.9492374  1.14361104  0.5275044
#>               hp         qsec         vs          wt
#>  [1,] -0.6969751 -0.403589576 -0.7844645 -0.46068061
#>  [2,] -0.9385480  0.728835293  1.2747549 -0.98466154
#>  [3,]  0.2266861 -0.403589576 -0.7844645  0.07274141
#>  [4,]  1.2213981 -1.244005517 -0.7844645  0.19547568
#>  [5,] -1.3790634  1.718816783  1.2747549 -0.16328603
#>  [6,] -0.5122429  0.508048055  1.2747549  0.07274141
#>  [7,]  0.2977369 -0.132947155 -0.7844645  0.66753057
#>  [8,]  0.2977369  0.009496225 -0.7844645  0.34653324
#>  [9,]  0.2977369  0.294382985 -0.7844645  0.39373873
#> [10,]  0.6529912  0.280138647 -0.7844645  1.78158010
#> [11,]  0.7950930  0.166183943 -0.7844645  1.94585520
#> [12,]  1.0082456 -0.118702817 -0.7844645  1.87127053
#> [13,] -1.3222227  1.341341827  1.2747549 -1.09795471
#> [14,] -1.5211651  0.664735773  1.2747549 -1.65025893
#> [15,] -0.8817074  1.725938952  1.2747549 -0.84776562
#> [16,] -0.1285682 -0.204168844 -0.7844645  0.06802086
#> [17,]  1.2213981 -1.550258784 -0.7844645  0.45038532
#> [18,] -0.6543446 -0.489055604  1.2747549 -1.74655812
#> [19,]  0.2266861 -1.486159263 -0.7844645 -0.55981214
#> [20,]  2.5003136 -2.127154472 -0.7844645  0.19547568
#> [21,] -0.7111853  0.721713124  1.2747549 -0.55037104
#> attr(,"center")
#>  [1]   0.3809524   3.1428571   6.3809524 242.4904762   3.5580952   3.6190476
#>  [7] 159.0476190  17.5866667   0.3809524   3.3629524
#> attr(,"scale")
#>  [1]   0.4856209   1.6700646   1.8120283 127.9874469   0.4825983   0.7221786
#>  [7]  70.3721239   1.4040667   0.4856209   1.0591989
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 47
#> 
#> $lambda.min
#> [1] 0.2145196
#> 
#> $null.dev
#> [1] 36.85673
#> 
#> $Bias
#> [1] 6.013551
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 13.95195