Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 45.488331 42.344815 38.246954 34.903615 30.532561 26.264053 22.729390
#>   [8] 19.821001 17.446133 15.524868 13.988400 12.777541 11.841423 11.136374
#>  [15] 10.624941 10.275045 10.029247  9.643844  9.329338  9.062964  8.903428
#>  [22]  8.586079  8.172212  7.824257  7.532815  7.289770  7.088115  6.921804
#>  [29]  6.785620  6.675069  6.586280  6.515924  7.653977  7.541531  7.438216
#>  [36]  7.350399  7.397779  7.594942  7.875865  8.141840  9.356113  9.486798
#>  [43]  9.487810  9.535010  9.555547  9.587342 10.870019 11.641661 11.137329
#>  [50] 11.204423 11.183140 11.130229 13.204572 13.022579 12.942093 12.874930
#>  [57] 12.815181 12.250335 12.479579 12.480075 12.743560 12.744676 14.635602
#>  [64] 14.636828 14.606504 14.176342 14.166295 14.166083 13.914356 14.562846
#>  [71] 14.976015 14.976008 15.654492 15.664882 15.680063 15.650799 15.650706
#>  [78] 15.650617 15.857479 15.857807 15.858127 15.858439 15.858745 15.859045
#>  [85] 15.859340 15.859622 15.853107 15.853366 15.645448 15.645470 15.645495
#>  [92] 15.645525 15.645559 15.645597 15.645640 15.645687 15.645738 15.645795
#>  [99] 15.645855 15.645920
#> 
#> $cvse
#>   [1] 13.830839 13.524536 13.079391 12.848063 11.703215 10.331800  9.158432
#>   [8]  8.148365  7.273084  6.509677  5.840191  5.251021  4.732324  4.277475
#>  [15]  3.882552  3.545796  3.271188  3.024901  2.807530  2.617127  2.586752
#>  [22]  2.537661  2.377130  2.239363  2.122270  2.023917  1.942477  1.876198
#>  [29]  1.823397  1.782456  1.751833  1.730079  1.868797  1.844309  1.821025
#>  [36]  1.799749  1.786776  1.802746  1.871673  1.941094  2.585236  2.657357
#>  [43]  2.676231  2.687968  2.693431  2.688695  2.826732  2.838751  2.851860
#>  [50]  2.839740  2.823957  2.828831  3.441553  3.426662  3.427014  3.379606
#>  [57]  3.338837  3.157474  3.214641  3.214783  3.281261  3.281622  3.603087
#>  [64]  3.603466  3.614211  3.529679  3.530131  3.530235  3.272624  3.647833
#>  [71]  3.652989  3.653200  3.793682  3.793484  3.790549  3.796615  3.796787
#>  [78]  3.796955  3.785330  3.785466  3.785601  3.785734  3.785867  3.786000
#>  [85]  3.786131  3.786264  3.787573  3.787709  3.747861  3.747960  3.748059
#>  [92]  3.748159  3.748259  3.748360  3.748461  3.748562  3.748664  3.748767
#>  [99]  3.748869  3.748972
#> 
#> $fold
#>  [1]  9  1 10  8  4  8  5  3  2  3  6  4  1  7 10  2  7  5  1  6  9
#> 
#> $lambda
#>   [1] 5.643217011 5.262883072 4.908182367 4.577387303 4.268886718 3.981178041
#>   [7] 3.712859966 3.462625630 3.229256252 3.011615189 2.808642404 2.619349306
#>  [13] 2.442813930 2.278176447 2.124634979 1.981441692 1.847899154 1.723356936
#>  [19] 1.607208446 1.498887976 1.397867943 1.303656322 1.215794249 1.133853786
#>  [25] 1.057435836 0.986168199 0.919703763 0.857718808 0.799911431 0.746000078
#>  [31] 0.695722169 0.648832824 0.605103663 0.564321701 0.526288307 0.490818237
#>  [37] 0.457738730 0.426888672 0.398117804 0.371285997 0.346262563 0.322925625
#>  [43] 0.301161519 0.280864240 0.261934931 0.244281393 0.227817645 0.212463498
#>  [49] 0.198144169 0.184789915 0.172335694 0.160720845 0.149888798 0.139786794
#>  [55] 0.130365632 0.121579425 0.113385379 0.105743584 0.098616820 0.091970376
#>  [61] 0.085771881 0.079991143 0.074600008 0.069572217 0.064883282 0.060510366
#>  [67] 0.056432170 0.052628831 0.049081824 0.045773873 0.042688867 0.039811780
#>  [73] 0.037128600 0.034626256 0.032292563 0.030116152 0.028086424 0.026193493
#>  [79] 0.024428139 0.022781764 0.021246350 0.019814417 0.018478992 0.017233569
#>  [85] 0.016072084 0.014988880 0.013978679 0.013036563 0.012157942 0.011338538
#>  [91] 0.010574358 0.009861682 0.009197038 0.008577188 0.007999114 0.007460001
#>  [97] 0.006957222 0.006488328 0.006051037 0.005643217
#> 
#> $fit
#> $beta
#>               5.6432     5.2629    4.9082    4.5774    4.2689    3.9812
#> (Intercept) 19.79524 21.7074542 23.490793 25.153941 26.704999 28.151520
#> am           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000 -0.5823417 -1.125436 -1.631927 -2.104282 -2.544802
#>                3.7129    3.4626    3.2293    3.0116   2.8086    2.6193
#> (Intercept) 29.500551 30.758661 31.931979 33.026220 34.04671 34.998426
#> am           0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> wt          -2.955633 -3.338775 -3.696094 -4.029331 -4.34011 -4.629943
#>                2.4428    2.2782    2.1246    1.9814    1.8479    1.7234
#> (Intercept) 35.885998 36.713751 37.485716 38.205653 38.710283 38.710283
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -4.900242 -5.152323 -5.387416 -5.606664 -5.760343 -5.760343
#>                1.6072      1.4989     1.3979     1.3037     1.2158     1.1339
#> (Intercept) 38.710283 36.89871389 34.8714116 32.9807420 31.2174973 29.5730893
#> am           0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.000000  0.09958947  0.2110387  0.3149766  0.4119095  0.5023095
#> vs           0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.760343 -5.74773775 -5.7336319 -5.7204768 -5.7082082 -5.6967665
#>                 1.0574    0.9862     0.9197     0.8577     0.7999     0.7460
#> (Intercept) 28.0395090 26.609287 25.2754568 24.0315225 22.8714252 21.7895146
#> am           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.5866167  0.665242  0.7385682  0.8069524  0.8707278  0.9302049
#> vs           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.6860960 -5.676145 -5.6668639 -5.6582087 -5.6501368 -5.6426089
#>                 0.6957    0.6488    0.6051   0.5643    0.5263    0.4908
#> (Intercept) 20.7805211 19.839530 18.961959 18.14353 17.536141 17.536101
#> am           0.0000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> carb         0.0000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> cyl          0.0000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> disp         0.0000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> drat         0.0000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> gear         0.0000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> hp           0.0000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> qsec         0.9856734  1.037404  1.085647  1.13064  1.164030  1.164033
#> vs           0.0000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> wt          -5.6355884 -5.629041 -5.622935 -5.61724 -5.613014 -5.613014
#>                0.4577    0.4269    0.3981     0.3713    0.3463    0.3229
#> (Intercept) 17.536101 17.536101 17.536101 15.3954849 12.570737 11.950488
#> am           0.000000  0.000000  0.000000  0.8315097  1.928467  2.169275
#> carb         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> qsec         1.164033  1.164033  1.164033  1.2109333  1.272849  1.286449
#> vs           0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> wt          -5.613014 -5.613014 -5.613014 -5.2994015 -4.885667 -4.794842
#>                0.3012    0.2809    0.2619    0.2443    0.2278    0.2125
#> (Intercept) 11.948868 11.947856 11.947223 11.946827 11.946580 11.946425
#> am           2.169845  2.170201  2.170424  2.170563  2.170650  2.170705
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         1.286489  1.286515  1.286530  1.286540  1.286547  1.286550
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -4.794626 -4.794491 -4.794407 -4.794354 -4.794321 -4.794300
#>                  0.1981     0.1848     0.1723     0.1607    0.1499       0.1398
#> (Intercept) 12.25590103 12.8196696 13.3471790  9.5096872  9.510706 -18.33475647
#> am           2.21061786  2.2873112  2.3594095  1.4210503  1.419859   0.00000000
#> carb        -0.03820615 -0.1088432 -0.1750251 -0.8544366 -0.854762   0.00000000
#> cyl          0.00000000  0.0000000  0.0000000  0.0000000  0.000000   0.00000000
#> disp         0.00000000  0.0000000  0.0000000  0.0000000  0.000000   0.02482366
#> drat         0.00000000  0.0000000  0.0000000  1.0506235  1.050353   1.40754315
#> gear         0.00000000  0.0000000  0.0000000  1.2487116  1.249721   2.32833368
#> hp           0.00000000  0.0000000  0.0000000  0.0000000  0.000000   0.00000000
#> qsec         1.26854170  1.2354614  1.2044855  0.8845918  0.884434   2.38997416
#> vs           0.00000000  0.0000000  0.0000000  0.0000000  0.000000  -2.19906138
#> wt          -4.76242648 -4.7023968 -4.6460615 -3.5782668 -3.578144  -6.91053479
#>                   0.1304       0.1216       0.1134       0.1057       0.0986
#> (Intercept) -18.33751405 -18.33997022 -18.34215802 -18.34410685 -18.34584288
#> am            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> carb          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> cyl           0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> disp          0.02482805   0.02483196   0.02483544   0.02483854   0.02484131
#> drat          1.40749563   1.40745324   1.40741543   1.40738171   1.40735165
#> gear          2.32855795   2.32875776   2.32893578   2.32909437   2.32923566
#> hp            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> qsec          2.39011181   2.39023441   2.39034363   2.39044093   2.39052762
#> vs           -2.19919176  -2.19930781  -2.19941113  -2.19950315  -2.19958510
#> wt           -6.91093501  -6.91129168  -6.91160953  -6.91189279  -6.91214521
#>                   0.0920       0.0858       0.0800       0.0746       0.0696
#> (Intercept) -18.34738942 -33.43552362 -33.44409251 -33.45216876 -33.45978070
#> am            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> carb          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> cyl           0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> disp          0.02484377   0.02407774   0.02407840   0.02407902   0.02407960
#> drat          1.40732484   1.53629732   1.53636285   1.53642462   1.53648283
#> gear          2.32936154   2.67041383   2.67067977   2.67093042   2.67116667
#> hp            0.00000000   0.02109755   0.02110713   0.02111615   0.02112466
#> qsec          2.39060485   3.12513218   3.12554388   3.12593192   3.12629764
#> vs           -2.19965811  -3.01596998  -3.01650636  -3.01701191  -3.01748839
#> wt           -6.91237015  -7.55938581  -7.55976021  -7.56011309  -7.56044569
#>                   0.0649       0.0605       0.0564       0.0526       0.0491
#> (Intercept) -33.03949829 -33.03132862 -33.02353194 -33.01609346 -33.00899884
#> am            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> carb         -0.30815498  -0.30821555  -0.30827130  -0.30832257  -0.30836968
#> cyl           0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> disp          0.02081906   0.02081728   0.02081563   0.02081408   0.02081263
#> drat          1.68451683   1.68449030   1.68446384   1.68443754   1.68441145
#> gear          2.85612541   2.85591505   2.85571287   2.85551865   2.85533217
#> hp            0.02522324   0.02521499   0.02520709   0.02519952   0.02519228
#> qsec          2.99981671   2.99939734   2.99899795   2.99861770   2.99825576
#> vs           -2.96582709  -2.96531777  -2.96483172  -2.96436801  -2.96392574
#> wt           -7.05269167  -7.05217927  -7.05169484  -7.05123694  -7.05080421
#>                   0.0458       0.0427       0.0398       0.0371       0.0346
#> (Intercept) -33.00223413 -32.99578583 -32.98964086 -32.98378655 -32.97821065
#> am            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> carb         -0.30841291  -0.30845256  -0.30848887  -0.30852209  -0.30855245
#> cyl           0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> disp          0.02081127   0.02081000   0.02080882   0.02080771   0.02080668
#> drat          1.68438564   1.68436016   1.68433507   1.68431040   1.68428618
#> gear          2.85515320   2.85498152   2.85481690   2.85465911   2.85450794
#> hp            0.02518535   0.02517872   0.02517238   0.02516632   0.02516053
#> qsec          2.99791133   2.99758367   2.99727201   2.99697566   2.99669393
#> vs           -2.96350406  -2.96310210  -2.96271906  -2.96235415  -2.96200660
#> wt           -7.05039534  -7.05000908  -7.04964427  -7.04929978  -7.04897452
#>                   0.0323       0.0301       0.0281       0.0262       0.0244
#> (Intercept) -32.97290134 -32.96784717 -42.41487329 -42.41939418 -42.42361630
#> am            0.00000000   0.00000000   0.52290053   0.52338813   0.52384778
#> carb         -0.30858015  -0.30860539  -0.52891865  -0.52892775  -0.52893020
#> cyl           0.00000000   0.00000000   0.58595008   0.58629206   0.58660930
#> disp          0.02080571   0.02080481   0.02009228   0.02009621   0.02010001
#> drat          1.68426246   1.68423925   2.01915294   2.01919475   2.01922767
#> gear          2.85436317   2.85422457   3.51967647   3.52006255   3.52041964
#> hp            0.02515500   0.02514972   0.02589276   0.02588927   0.02588595
#> qsec          2.99642616   2.99617172   3.10019746   3.10023449   3.10027212
#> vs           -2.96167568  -2.96136066  -2.46382899  -2.46324382  -2.46269605
#> wt           -7.04866750  -7.04837775  -6.86099287  -6.86119132  -6.86139209
#>                   0.0228       0.0212       0.0198       0.0185       0.0172
#> (Intercept) -42.42755374 -42.43122011 -42.43462854 -42.43779168 -42.44072172
#> am            0.52428062   0.52468778   0.52507034   0.52542936   0.52576585
#> carb         -0.52892648  -0.52891707  -0.52890242  -0.52888296  -0.52885908
#> cyl           0.58690301   0.58717436   0.58742447   0.58765441   0.58786524
#> disp          0.02010369   0.02010725   0.02011070   0.02011403   0.02011725
#> drat          2.01925232   2.01926927   2.01927906   2.01928222   2.01927925
#> gear          3.52074920   3.52105263   3.52133128   3.52158646   3.52181941
#> hp            0.02588281   0.02587983   0.02587701   0.02587435   0.02587183
#> qsec          3.10031024   3.10034875   3.10038756   3.10042657   3.10046571
#> vs           -2.46218401  -2.46170611  -2.46126076  -2.46084647  -2.46046176
#> wt           -6.86159461  -6.86179833  -6.86200277  -6.86220746  -6.86241196
#>                   0.0161       0.0150       0.0140       0.0130       0.0122
#> (Intercept) -42.44343042 -42.44592907 -42.44822856 -42.45033935 -42.45227150
#> am            0.52608082   0.52637523   0.52665000   0.52690604   0.52714423
#> carb         -0.52883118  -0.52879961  -0.52876471  -0.52872680  -0.52868620
#> cyl           0.58805795   0.58823351   0.58839286   0.58853688   0.58866643
#> disp          0.02012036   0.02012336   0.02012625   0.02012903   0.02013172
#> drat          2.01927061   2.01925676   2.01923811   2.01921507   2.01918803
#> gear          3.52203134   3.52222340   3.52239670   3.52255230   3.52269123
#> hp            0.02586946   0.02586722   0.02586511   0.02586313   0.02586127
#> qsec          3.10050488   3.10054403   3.10058309   3.10062198   3.10066065
#> vs           -2.46010521  -2.45977547  -2.45947121  -2.45919116  -2.45893408
#> wt           -6.86261590  -6.86281889  -6.86302061  -6.86322073  -6.86341899
#>                   0.0113       0.0106       0.0099       0.0092       0.0086
#> (Intercept) -42.45403465 -42.45563808 -42.45709068 -42.45840099 -42.45957718
#> am            0.52736542   0.52757042   0.52776001   0.52793497   0.52809603
#> carb         -0.52864318  -0.52859803  -0.52855098  -0.52850230  -0.52845218
#> cyl           0.58878233   0.58888537   0.58897630   0.58905586   0.58912472
#> disp          0.02013430   0.02013679   0.02013918   0.02014147   0.02014368
#> drat          2.01915733   2.01912333   2.01908634   2.01904667   2.01900461
#> gear          3.52281445   3.52292291   3.52301750   3.52309907   3.52316845
#> hp            0.02585952   0.02585789   0.02585636   0.02585493   0.02585360
#> qsec          3.10069905   3.10073713   3.10077483   3.10081213   3.10084898
#> vs           -2.45869879  -2.45848416  -2.45828907  -2.45811247  -2.45795335
#> wt           -6.86361511  -6.86380886  -6.86400003  -6.86418843  -6.86437389
#>                   0.0080       0.0075       0.0070       0.0065       0.0061
#> (Intercept) -42.46062707 -42.46155817 -42.46237764 -42.46309233 -42.46370881
#> am            0.52824390   0.52837926   0.52850276   0.52861505   0.52871673
#> carb         -0.52840086  -0.52834852  -0.52829536  -0.52824154  -0.52818722
#> cyl           0.58918357   0.58923302   0.58927369   0.58930615   0.58933096
#> disp          0.02014579   0.02014782   0.02014977   0.02015163   0.02015341
#> drat          2.01896042   2.01891436   2.01886667   2.01881756   2.01876726
#> gear          3.52322641   3.52327370   3.52331103   3.52333909   3.52335851
#> hp            0.02585236   0.02585121   0.02585014   0.02584916   0.02584825
#> qsec          3.10088534   3.10092119   3.10095649   3.10099122   3.10102536
#> vs           -2.45781074  -2.45768368  -2.45757129  -2.45747270  -2.45738708
#> wt           -6.86455624  -6.86473536  -6.86491112  -6.86508342  -6.86525217
#>                   0.0056
#> (Intercept) -42.46423331
#> am            0.52880838
#> carb         -0.52813256
#> cyl           0.58934863
#> disp          0.02015512
#> drat          2.01871596
#> gear          3.52336991
#> hp            0.02584741
#> qsec          3.10105888
#> vs           -2.45731364
#> wt           -6.86541730
#> 
#> $iter
#>   [1]   0   3   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   4   3   3   3   3   3   3   3   3   3   3   3   3   3   3   2   1
#>  [37]   1   1   1  54  57  11   1   1   1   1   1   1  17  19  19  66   2 191
#>  [55]   1   1   1   1   1   1  82   1   1   1 101   1   1   1   1   1   1   1
#>  [73]   1   1   1   1 103   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.643217011 5.262883072 4.908182367 4.577387303 4.268886718 3.981178041
#>   [7] 3.712859966 3.462625630 3.229256252 3.011615189 2.808642404 2.619349306
#>  [13] 2.442813930 2.278176447 2.124634979 1.981441692 1.847899154 1.723356936
#>  [19] 1.607208446 1.498887976 1.397867943 1.303656322 1.215794249 1.133853786
#>  [25] 1.057435836 0.986168199 0.919703763 0.857718808 0.799911431 0.746000078
#>  [31] 0.695722169 0.648832824 0.605103663 0.564321701 0.526288307 0.490818237
#>  [37] 0.457738730 0.426888672 0.398117804 0.371285997 0.346262563 0.322925625
#>  [43] 0.301161519 0.280864240 0.261934931 0.244281393 0.227817645 0.212463498
#>  [49] 0.198144169 0.184789915 0.172335694 0.160720845 0.149888798 0.139786794
#>  [55] 0.130365632 0.121579425 0.113385379 0.105743584 0.098616820 0.091970376
#>  [61] 0.085771881 0.079991143 0.074600008 0.069572217 0.064883282 0.060510366
#>  [67] 0.056432170 0.052628831 0.049081824 0.045773873 0.042688867 0.039811780
#>  [73] 0.037128600 0.034626256 0.032292563 0.030116152 0.028086424 0.026193493
#>  [79] 0.024428139 0.022781764 0.021246350 0.019814417 0.018478992 0.017233569
#>  [85] 0.016072084 0.014988880 0.013978679 0.013036563 0.012157942 0.011338538
#>  [91] 0.010574358 0.009861682 0.009197038 0.008577188 0.007999114 0.007460001
#>  [97] 0.006957222 0.006488328 0.006051037 0.005643217
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 39
#> 
#> $loss
#>   [1] 803.98952 675.60708 568.19608 478.73857 404.62898 343.61914 293.77048
#>   [8] 253.41260 221.10725 195.61717 175.87891 160.97941 150.13557 142.67659
#>  [15] 138.02866 135.70166 135.22566 135.22566 135.22566 126.46809 117.59573
#>  [22] 110.20473 104.08020  99.03669  94.91432  91.57529  88.90097  86.78930
#>  [29]  85.15253  83.91536  83.01316  82.39058  82.00029  81.80184  81.75784
#>  [36]  81.75784  81.75784  81.75784  81.75784  76.55835  73.46758  73.36297
#>  [43]  73.36297  73.36296  73.36296  73.36296  73.36296  73.36296  72.95990
#>  [50]  72.34443  71.92060  64.94962  64.94961  53.57680  53.57679  53.57678
#>  [57]  53.57677  53.57677  53.57677  53.57676  51.29962  51.29960  51.29958
#>  [64]  51.29957  50.73787  50.73785  50.73784  50.73782  50.73781  50.73779
#>  [71]  50.73778  50.73777  50.73776  50.73775  50.73774  50.73774  50.07555
#>  [78]  50.07553  50.07552  50.07550  50.07549  50.07548  50.07547  50.07546
#>  [85]  50.07545  50.07545  50.07544  50.07543  50.07543  50.07542  50.07542
#>  [92]  50.07542  50.07541  50.07541  50.07541  50.07540  50.07540  50.07540
#>  [99]  50.07540  50.07540
#> 
#> $linear.predictors
#>      5.6432   5.2629   4.9082   4.5774   4.2689   3.9812   3.7129   3.4626
#> 1  19.79524 20.03322 20.25517 20.46215 20.65519 20.83521 21.00311 21.15968
#> 2  19.79524 19.83523 19.87252 19.90730 19.93973 19.96998 19.99819 20.02450
#> 3  19.79524 19.69255 19.59679 19.50748 19.42418 19.34650 19.27406 19.20650
#> 4  19.79524 19.62849 19.47299 19.32796 19.19271 19.06658 18.94894 18.83924
#> 5  19.79524 19.84978 19.90065 19.94810 19.99234 20.03360 20.07208 20.10797
#> 6  19.79524 19.70420 19.61930 19.54011 19.46627 19.39740 19.33317 19.27328
#> 7  19.79524 19.70420 19.61930 19.54011 19.46627 19.39740 19.33317 19.27328
#> 8  19.79524 19.33732 18.91027 18.51200 18.14057 17.79418 17.47113 17.16985
#> 9  19.79524 19.53532 19.29292 19.06685 18.85603 18.65941 18.47604 18.30503
#> 10 19.79524 18.65016 17.58226 16.58633 15.65752 14.79131 13.98348 13.23009
#> 11 19.79524 18.54883 17.38643 16.30237 15.29137 14.34851 13.46920 12.64915
#> 12 19.79524 20.76697 21.67321 22.51838 23.30658 24.04166 24.72720 25.36654
#> 13 19.79524 20.63886 21.42562 22.15936 22.84364 23.48181 24.07696 24.63201
#> 14 19.79524 19.65761 19.52926 19.40956 19.29793 19.19382 19.09672 19.00617
#> 15 19.79524 19.70711 19.62492 19.54827 19.47679 19.41012 19.34795 19.28997
#> 16 19.79524 19.47126 19.16912 18.88734 18.62456 18.37948 18.15092 17.93777
#> 17 19.79524 19.46835 19.16349 18.87918 18.61403 18.36676 18.13614 17.92107
#> 18 19.79524 20.46124 21.08236 21.66162 22.20184 22.70564 23.17550 23.61368
#> 19 19.79524 20.82637 21.78801 22.68484 23.52122 24.30123 25.02868 25.70710
#> 20 19.79524 20.09437 20.37334 20.63350 20.87614 21.10242 21.31345 21.51026
#> 21 19.79524 20.08854 20.36208 20.61719 20.85509 21.07697 21.28389 21.47687
#>      3.2293   3.0116   2.8086    2.6193    2.4428    2.2782    2.1246    1.9814
#> 1  21.30571 21.44189 21.56890 21.687341 21.797804 21.900821 21.996895 22.086495
#> 2  20.04904 20.07192 20.09326 20.113161 20.131721 20.149031 20.165174 20.180229
#> 3  19.14349 19.08473 19.02993 18.978825 18.931162 18.886712 18.845257 18.806596
#> 4  18.73692 18.64151 18.55252 18.469531 18.392136 18.319956 18.252641 18.189863
#> 5  20.14144 20.17265 20.20176 20.228909 20.254227 20.277839 20.299859 20.320396
#> 6  19.21742 19.16532 19.11673 19.071424 19.029167 18.989758 18.953005 18.918730
#> 7  19.21742 19.16532 19.11673 19.071424 19.029167 18.989758 18.953005 18.918730
#> 8  16.88888 16.62684 16.38247 16.154560 15.942015 15.743794 15.558933 15.386532
#> 9  18.14555 17.99681 17.85810 17.728740 17.608097 17.495584 17.390655 17.292797
#> 10 12.52749 11.87223 11.26114 10.691228 10.159730  9.664053  9.201783  8.770668
#> 11 11.88436 11.17113 10.50596  9.885618  9.307088  8.767548  8.264372  7.795109
#> 12 25.96279 26.51885 27.03743 27.521069 27.972108 28.392749 28.785039 29.150891
#> 13 25.14965 25.63240 26.08261 26.502482 26.894055 27.259237 27.599808 27.917425
#> 14 18.92173 18.84297 18.76953 18.701028 18.637148 18.577572 18.522012 18.470197
#> 15 19.23590 19.18547 19.13843 19.094573 19.053668 19.015520 18.979943 18.946763
#> 16 17.73898 17.55359 17.38069 17.219447 17.069070 16.928829 16.798039 16.676064
#> 17 17.72050 17.53344 17.35899 17.196297 17.044569 16.903067 16.771102 16.648031
#> 18 24.02234 24.40345 24.75888 25.090349 25.399481 25.687779 25.956646 26.207393
#> 19 26.33979 26.92984 27.48013 27.993323 28.471933 28.918286 29.334556 29.722771
#> 20 21.69380 21.86497 22.02461 22.173485 22.312329 22.441815 22.562574 22.675194
#> 21 21.65684 21.82468 21.98121 22.127186 22.263326 22.390292 22.508700 22.619128
#>       1.8479    1.7234    1.6072    1.4989    1.3979    1.3037    1.2158
#> 1  22.149298 22.149298 22.149298 22.068981 21.979098 21.895274 21.817099
#> 2  20.190782 20.190782 20.190782 20.355756 20.540377 20.712555 20.873129
#> 3  18.779498 18.779498 18.779498 19.025240 19.300248 19.556720 19.795908
#> 4  18.145860 18.145860 18.145860 17.956787 17.745199 17.547870 17.363841
#> 5  20.334790 20.334790 20.334790 20.555220 20.801900 21.031954 21.246504
#> 6  18.894705 18.894705 18.894705 18.948983 19.009726 19.066375 19.119205
#> 7  18.894705 18.894705 18.894705 19.008737 19.136349 19.255361 19.366351
#> 8  15.265689 15.265689 15.265689 15.238278 15.207603 15.178995 15.152316
#> 9  17.224205 17.224205 17.224205 17.212427 17.199246 17.186953 17.175488
#> 10  8.468485  8.468485  8.468485  8.513709  8.564320  8.611519  8.655538
#> 11  7.466185  7.466185  7.466185  7.497669  7.532902  7.565760  7.596404
#> 12 29.407330 29.407330 29.407330 29.460514 29.520033 29.575540 29.627306
#> 13 28.140054 28.140054 28.140054 28.333446 28.549867 28.751702 28.939935
#> 14 18.433877 18.433877 18.433877 18.346751 18.249250 18.158320 18.073518
#> 15 18.923506 18.923506 18.923506 18.878132 18.827355 18.780000 18.735837
#> 16 16.590568 16.590568 16.590568 16.362075 16.106371 15.867901 15.645504
#> 17 16.561766 16.561766 16.561766 16.496663 16.423807 16.355861 16.292494
#> 18 26.383150 26.383150 26.383150 26.261699 26.125785 25.999032 25.880821
#> 19 29.994885 29.994885 29.994885 29.885449 29.762980 29.648766 29.542249
#> 20 22.754134 22.754134 22.754134 22.521117 22.260351 22.017159 21.790358
#> 21 22.696531 22.696531 22.696531 22.772367 22.857235 22.936382 23.010196
#>       1.1339    1.0574    0.9862    0.9197    0.8577    0.7999    0.7460
#> 1  21.744193 21.676200 21.612790 21.553653 21.498502 21.447069 21.399101
#> 2  21.022881 21.162540 21.292786 21.414255 21.527536 21.633183 21.731710
#> 3  20.018975 20.227007 20.421020 20.601956 20.770698 20.928067 21.074830
#> 4  17.192215 17.032156 16.882884 16.743673 16.613844 16.492765 16.379846
#> 5  21.446593 21.633198 21.807225 21.969524 22.120885 22.262044 22.393690
#> 6  19.168476 19.214425 19.257278 19.297243 19.334514 19.369273 19.401689
#> 7  19.469861 19.566395 19.656423 19.740383 19.818685 19.891709 19.959812
#> 8  15.127434 15.104230 15.082589 15.062407 15.043585 15.026032 15.009661
#> 9  17.164797 17.154826 17.145527 17.136854 17.128766 17.121224 17.114189
#> 10  8.696589  8.734874  8.770579  8.803877  8.834931  8.863892  8.890902
#> 11  7.624982  7.651635  7.676491  7.699672  7.721290  7.741452  7.760255
#> 12 29.675583 29.720606 29.762595 29.801754 29.838274 29.872333 29.904096
#> 13 29.115481 29.279196 29.431877 29.574268 29.707062 29.830907 29.946404
#> 14 17.994432 17.920676 17.851890 17.787741 17.727915 17.672121 17.620088
#> 15 18.694650 18.656239 18.620417 18.587009 18.555852 18.526796 18.499697
#> 16 15.438095 15.244664 15.064271 14.896035 14.739138 14.592815 14.456354
#> 17 16.233398 16.178285 16.126887 16.078952 16.034249 15.992558 15.953677
#> 18 25.770577 25.667763 25.571879 25.482457 25.399061 25.321286 25.248753
#> 19 29.442912 29.350269 29.263870 29.183294 29.108148 29.038068 28.972710
#> 20 21.578843 21.381583 21.197617 21.026050 20.866047 20.716827 20.577664
#> 21 23.079034 23.143234 23.203106 23.258943 23.311017 23.359581 23.404873
#>      0.6957    0.6488    0.6051    0.5643    0.5263    0.4908    0.4577
#> 1  21.35437 21.312647 21.273739 21.237453 21.210524 21.210522 21.210522
#> 2  21.82360 21.909290 21.989208 22.063739 22.119053 22.119057 22.119057
#> 3  21.21170 21.339349 21.458393 21.569414 21.651808 21.651814 21.651814
#> 4  16.27454 16.176327 16.084735 15.999317 15.935923 15.935919 15.935919
#> 5  22.51646 22.630962 22.737743 22.837329 22.911235 22.911240 22.911240
#> 6  19.43192 19.460115 19.486409 19.510931 19.529130 19.529131 19.529131
#> 7  20.02333 20.082557 20.137798 20.189315 20.227548 20.227551 20.227551
#> 8  14.99439 14.980156 14.966878 14.954494 14.945304 14.945303 14.945303
#> 9  17.10763 17.101511 17.095805 17.090484 17.086535 17.086534 17.086534
#> 10  8.91609  8.939582  8.961490  8.981921  8.997084  8.997085  8.997085
#> 11  7.77779  7.794144  7.809396  7.823619  7.834175  7.834176  7.834176
#> 12 29.93372 29.961344 29.987108 30.011136 30.028968 30.028969 30.028969
#> 13 30.05412 30.154572 30.248256 30.335625 30.400467 30.400471 30.400471
#> 14 17.57156 17.526305 17.484099 17.444737 17.415525 17.415523 17.415523
#> 15 18.47443 18.450857 18.428877 18.408378 18.393165 18.393164 18.393164
#> 16 14.32909 14.210402 14.099714 13.996487 13.919876 13.919871 13.919871
#> 17 15.91742 15.883599 15.852061 15.822649 15.800821 15.800820 15.800820
#> 18 25.18111 25.118023 25.059189 25.004320 24.963600 24.963597 24.963597
#> 19 28.91176 28.854912 28.801899 28.752458 28.715766 28.715763 28.715763
#> 20 20.44788 20.326843 20.213963 20.108691 20.030564 20.030559 20.030559
#> 21 23.44711 23.486503 23.523241 23.557502 23.582929 23.582930 23.582930
#>       0.4269    0.3981    0.3713    0.3463    0.3229    0.3012    0.2809
#> 1  21.210522 21.210522 21.601300 22.116795 22.229951 22.230212 22.230375
#> 2  22.119057 22.119057 21.898453 21.607494 21.543636 21.543498 21.543413
#> 3  21.651814 21.651814 21.544627 21.403328 21.372329 21.372277 21.372244
#> 4  15.935919 15.935919 15.657805 15.290827 15.210251 15.210045 15.209916
#> 5  22.911240 22.911240 22.709060 22.442431 22.383918 22.383798 22.383723
#> 6  19.529131 19.529131 19.325623 19.057172 18.998245 18.998110 18.998025
#> 7  20.227551 20.227551 20.052183 19.820881 19.770114 19.770003 19.769934
#> 8  14.945303 14.945303 14.897160 14.833637 14.819690 14.819655 14.819633
#> 9  17.086534 17.086534 16.941144 16.749334 16.707226 16.707126 16.707063
#> 10  8.997085  8.997085  9.346208  9.806802  9.907917  9.908160  9.908312
#> 11  7.834176  7.834176  8.230363  8.753040  8.867783  8.868057  8.868228
#> 12 30.028969 30.028969 30.094946 30.182008 30.201125 30.201175 30.201206
#> 13 30.400471 30.400471 30.600166 30.863693 30.921559 30.921712 30.921808
#> 14 17.415523 17.415523 17.170037 16.846145 16.775035 16.774860 16.774750
#> 15 18.393164 18.393164 18.141187 17.808751 17.735770 17.735594 17.735483
#> 16 13.919871 13.919871 13.706265 13.424372 13.362471 13.362305 13.362202
#> 17 15.800820 15.800820 15.665699 15.487416 15.448273 15.448175 15.448113
#> 18 24.963597 24.963597 25.108862 25.300449 25.342496 25.342586 25.342641
#> 19 28.715763 28.715763 28.673773 28.618332 28.606152 28.606114 28.606090
#> 20 20.030559 20.030559 20.317119 20.695060 20.778007 20.778184 20.778295
#> 21 23.582930 23.582930 24.018018 24.592034 24.718050 24.718355 24.718545
#>       0.2619    0.2443    0.2278    0.2125    0.1981    0.1848    0.1723
#> 1  22.230477 22.230541 22.230581 22.230606 22.212298 22.179771 22.149405
#> 2  21.543359 21.543325 21.543304 21.543291 21.566944 21.609991 21.650265
#> 3  21.372223 21.372210 21.372202 21.372197 21.389612 21.421563 21.451479
#> 4  15.209835 15.209784 15.209753 15.209733 15.194914 15.166449 15.139690
#> 5  22.383676 22.383647 22.383628 22.383617 22.358182 22.310566 22.265903
#> 6  18.997973 18.997940 18.997919 18.997906 18.934642 18.816996 18.706712
#> 7  19.769891 19.769864 19.769847 19.769837 19.695767 19.558273 19.429404
#> 8  14.819619 14.819611 14.819605 14.819602 14.830832 14.851414 14.870682
#> 9  16.707023 16.706999 16.706984 16.706974 16.703766 16.697321 16.691240
#> 10  9.908407  9.908467  9.908504  9.908527  9.908717  9.910310  9.911906
#> 11  8.868336  8.868403  8.868444  8.868471  8.877088  8.894419  8.910773
#> 12 30.201225 30.201237 30.201245 30.201250 30.192180 30.175669 30.160221
#> 13 30.921868 30.921905 30.921929 30.921943 30.933240 30.954922 30.975303
#> 14 16.774682 16.774639 16.774612 16.774595 16.816046 16.891781 16.962663
#> 15 17.735414 17.735371 17.735344 17.735327 17.766325 17.822733 17.875507
#> 16 13.362137 13.362097 13.362071 13.362055 13.363586 13.365554 13.367325
#> 17 15.448075 15.448051 15.448036 15.448027 15.496595 15.585885 15.669501
#> 18 25.342676 25.342698 25.342712 25.342720 25.383160 25.458371 25.528875
#> 19 28.606075 28.606066 28.606060 28.606057 28.622910 28.653866 28.682853
#> 20 20.778364 20.778407 20.778434 20.778451 20.707757 20.577934 20.456374
#> 21 24.718664 24.718739 24.718785 24.718814 24.745437 24.796214 24.843918
#>       0.1607    0.1499    0.1398   0.1304    0.1216    0.1134    0.1057
#> 1  21.373505 21.372679 21.249355 21.24920 21.249068 21.248947 21.248839
#> 2  21.329643 21.329857 21.434650 21.43481 21.434951 21.435077 21.435190
#> 3  20.806749 20.806957 20.336154 20.33619 20.336227 20.336258 20.336286
#> 4  14.448099 14.447913 15.291559 15.29165 15.291735 15.291809 15.291875
#> 5  22.949626 22.950268 23.369860 23.36981 23.369771 23.369734 23.369701
#> 6  19.084024 19.084251 18.421820 18.42152 18.421252 18.421014 18.420802
#> 7  19.614779 19.614911 19.855804 19.85559 19.855393 19.855220 19.855066
#> 8  14.746278 14.746271 13.277442 13.27719 13.276960 13.276758 13.276577
#> 9  16.139807 16.139727 16.105019 16.10493 16.104846 16.104774 16.104709
#> 10 10.035462 10.035221 11.182543 11.18276 11.182958 11.183133 11.183288
#> 11  9.344853  9.344639  9.398358  9.39843  9.398494  9.398552  9.398603
#> 12 30.000024 29.999179 30.699664 30.69967 30.699682 30.699690 30.699696
#> 13 30.542036 30.541518 31.363966 31.36409 31.364202 31.364302 31.364390
#> 14 16.774234 16.774653 16.422770 16.42286 16.422944 16.423017 16.423083
#> 15 17.868504 17.868739 18.239265 18.23937 18.239464 18.239548 18.239622
#> 16 13.647916 13.647691 12.881711 12.88157 12.881441 12.881328 12.881227
#> 17 16.106723 16.107067 17.092996 17.09333 17.093622 17.093885 17.094120
#> 18 27.234876 27.235528 27.652639 27.65276 27.652871 27.652969 27.653056
#> 19 28.961957 28.962679 28.709943 28.71013 28.710306 28.710458 28.710594
#> 20 19.650307 19.650143 19.904068 19.90392 19.903789 19.903671 19.903567
#> 21 25.040599 25.040107 22.810415 22.81021 22.810022 22.809857 22.809710
#>       0.0986    0.0920    0.0858    0.0800    0.0746    0.0696    0.0649
#> 1  21.248743 21.248658 20.867375 20.867215 20.867064 20.866922 20.607996
#> 2  21.435290 21.435379 21.273475 21.273392 21.273314 21.273240 21.231114
#> 3  20.336310 20.336332 20.467360 20.467416 20.467469 20.467519 20.490871
#> 4  15.291933 15.291986 15.859204 15.859411 15.859606 15.859789 15.717099
#> 5  23.369671 23.369645 23.127554 23.127484 23.127418 23.127356 23.134978
#> 6  18.420612 18.420444 18.068506 18.068256 18.068020 18.067798 18.016981
#> 7  19.854929 19.854807 19.943586 19.943582 19.943579 19.943576 19.816871
#> 8  13.276416 13.276273 13.340949 13.340924 13.340901 13.340878 13.550314
#> 9  16.104651 16.104600 16.536167 16.536352 16.536525 16.536689 16.548193
#> 10 11.183427 11.183550 11.269860 11.269991 11.270115 11.270231 11.139324
#> 11  9.398648  9.398688  9.484089  9.484182  9.484269  9.484351  9.552505
#> 12 30.699702 30.699708 30.392905 30.392754 30.392612 30.392478 30.161654
#> 13 31.364469 31.364539 32.115262 32.115672 32.116058 32.116423 32.094091
#> 14 16.423141 16.423193 15.749193 15.748876 15.748577 15.748296 15.747214
#> 15 18.239689 18.239748 17.997616 17.997524 17.997437 17.997355 18.002108
#> 16 12.881137 12.881057 13.032460 13.032416 13.032375 13.032336 13.190709
#> 17 17.094329 17.094515 16.848345 16.848295 16.848247 16.848202 16.871845
#> 18 27.653133 27.653202 27.551393 27.551469 27.551541 27.551608 27.891255
#> 19 28.710716 28.710824 28.743616 28.743623 28.743630 28.743636 28.865918
#> 20 19.903474 19.903391 20.161335 20.161449 20.161555 20.161656 19.884183
#> 21 22.809578 22.809462 22.869749 22.869717 22.869687 22.869658 23.184777
#>      0.0605    0.0564    0.0526    0.0491   0.0458    0.0427    0.0398
#> 1  20.60812 20.608246 20.608364 20.608477 20.60859 20.608692 20.608793
#> 2  21.23115 21.231185 21.231220 21.231254 21.23129 21.231320 21.231352
#> 3  20.49081 20.490759 20.490707 20.490657 20.49061 20.490564 20.490521
#> 4  15.71684 15.716590 15.716356 15.716134 15.71592 15.715724 15.715535
#> 5  23.13507 23.135161 23.135246 23.135326 23.13540 23.135474 23.135542
#> 6  18.01725 18.017504 18.017747 18.017979 18.01820 18.018408 18.018607
#> 7  19.81689 19.816903 19.816918 19.816932 19.81695 19.816958 19.816971
#> 8  13.55040 13.550489 13.550568 13.550641 13.55071 13.550773 13.550832
#> 9  16.54803 16.547865 16.547712 16.547566 16.54743 16.547293 16.547166
#> 10 11.13917 11.139015 11.138873 11.138737 11.13861 11.138488 11.138373
#> 11  9.55244  9.552376  9.552315  9.552257  9.55220  9.552146  9.552093
#> 12 30.16174 30.161818 30.161898 30.161975 30.16205 30.162124 30.162195
#> 13 32.09369 32.093305 32.092939 32.092590 32.09226 32.091942 32.091640
#> 14 15.74749 15.747747 15.747996 15.748235 15.74846 15.748680 15.748888
#> 15 18.00217 18.002232 18.002290 18.002347 18.00240 18.002452 18.002502
#> 16 13.19077 13.190828 13.190883 13.190934 13.19098 13.191028 13.191070
#> 17 16.87185 16.871853 16.871858 16.871863 16.87187 16.871876 16.871883
#> 18 27.89126 27.891269 27.891271 27.891272 27.89127 27.891266 27.891261
#> 19 28.86591 28.865905 28.865898 28.865892 28.86589 28.865879 28.865873
#> 20 19.88406 19.883944 19.883834 19.883730 19.88363 19.883538 19.883450
#> 21 23.18490 23.185005 23.185108 23.185203 23.18529 23.185375 23.185451
#>       0.0371    0.0346    0.0323    0.0301    0.0281    0.0262   0.0244
#> 1  20.608890 20.608983 20.609072 20.609158 20.564431 20.564423 20.56442
#> 2  21.231383 21.231413 21.231442 21.231471 21.127859 21.127963 21.12806
#> 3  20.490480 20.490441 20.490404 20.490369 20.426432 20.426390 20.42635
#> 4  15.715357 15.715188 15.715028 15.714877 15.387895 15.387800 15.38772
#> 5  23.135607 23.135668 23.135725 23.135780 22.606912 22.606491 22.60609
#> 6  18.018797 18.018977 18.019148 18.019311 18.199183 18.199194 18.19920
#> 7  19.816982 19.816993 19.817004 19.817014 20.059302 20.059335 20.05936
#> 8  13.550888 13.550939 13.550987 13.551031 13.665144 13.664908 13.66468
#> 9  16.547045 16.546929 16.546819 16.546714 16.617921 16.617759 16.61761
#> 10 11.138265 11.138162 11.138065 11.137973 11.145112 11.145330 11.14554
#> 11  9.552043  9.551995  9.551948  9.551904  9.614428  9.614527  9.61462
#> 12 30.162264 30.162331 30.162396 30.162459 30.165854 30.165987 30.16611
#> 13 32.091353 32.091080 32.090820 32.090572 32.274209 32.274265 32.27432
#> 14 15.749087 15.749277 15.749458 15.749630 15.769678 15.769798 15.76991
#> 15 18.002550 18.002596 18.002640 18.002682 18.192125 18.192239 18.19234
#> 16 13.191111 13.191148 13.191184 13.191217 13.051378 13.051196 13.05102
#> 17 16.871890 16.871897 16.871905 16.871913 17.038906 17.039216 17.03951
#> 18 27.891254 27.891246 27.891237 27.891227 27.801336 27.801116 27.80090
#> 19 28.865866 28.865860 28.865854 28.865848 28.990063 28.990357 28.99063
#> 20 19.883366 19.883287 19.883213 19.883143 19.850656 19.850684 19.85071
#> 21 23.185523 23.185589 23.185651 23.185708 23.151175 23.151024 23.15088
#>       0.0228    0.0212   0.0198    0.0185    0.0172   0.0161    0.0150
#> 1  20.564415 20.564413 20.56441 20.564416 20.564419 20.56442 20.564430
#> 2  21.128163 21.128260 21.12836 21.128447 21.128538 21.12863 21.128711
#> 3  20.426318 20.426287 20.42626 20.426234 20.426212 20.42619 20.426176
#> 4  15.387649 15.387592 15.38755 15.387509 15.387483 15.38747 15.387456
#> 5  22.605722 22.605373 22.60505 22.604738 22.604451 22.60418 22.603934
#> 6  18.199196 18.199188 18.19918 18.199157 18.199135 18.19911 18.199079
#> 7  20.059382 20.059397 20.05941 20.059413 20.059415 20.05941 20.059406
#> 8  13.664454 13.664237 13.66403 13.663821 13.663622 13.66343 13.663243
#> 9  16.617458 16.617318 16.61718 16.617056 16.616935 16.61682 16.616710
#> 10 11.145744 11.145940 11.14613 11.146312 11.146488 11.14666 11.146819
#> 11  9.614709  9.614792  9.61487  9.614945  9.615014  9.61508  9.615142
#> 12 30.166238 30.166358 30.16647 30.166583 30.166690 30.16679 30.166891
#> 13 32.274372 32.274424 32.27447 32.274521 32.274567 32.27461 32.274656
#> 14 15.770016 15.770115 15.77021 15.770296 15.770377 15.77045 15.770524
#> 15 18.192442 18.192532 18.19261 18.192690 18.192760 18.19282 18.192881
#> 16 13.050860 13.050704 13.05056 13.050416 13.050284 13.05016 13.050040
#> 17 17.039787 17.040050 17.04030 17.040534 17.040757 17.04097 17.041163
#> 18 27.800698 27.800500 27.80031 27.800125 27.799948 27.79978 27.799615
#> 19 28.990898 28.991147 28.99138 28.991604 28.991813 28.99201 28.992196
#> 20 19.850744 19.850776 19.85081 19.850843 19.850878 19.85091 19.850950
#> 21 23.150736 23.150598 23.15047 23.150338 23.150214 23.15009 23.149980
#>       0.0140    0.0130    0.0122    0.0113    0.0106    0.0099    0.0092
#> 1  20.564437 20.564445 20.564454 20.564464 20.564474 20.564485 20.564497
#> 2  21.128794 21.128875 21.128954 21.129030 21.129104 21.129176 21.129245
#> 3  20.426161 20.426148 20.426138 20.426129 20.426123 20.426118 20.426114
#> 4  15.387454 15.387459 15.387471 15.387489 15.387512 15.387541 15.387574
#> 5  22.603701 22.603485 22.603285 22.603099 22.602928 22.602770 22.602625
#> 6  18.199046 18.199010 18.198971 18.198929 18.198886 18.198840 18.198793
#> 7  20.059396 20.059383 20.059367 20.059349 20.059328 20.059305 20.059280
#> 8  13.663062 13.662887 13.662719 13.662556 13.662398 13.662246 13.662100
#> 9  16.616606 16.616507 16.616413 16.616324 16.616240 16.616161 16.616086
#> 10 11.146976 11.147126 11.147270 11.147408 11.147541 11.147668 11.147790
#> 11  9.615199  9.615254  9.615304  9.615352  9.615396  9.615438  9.615476
#> 12 30.166986 30.167077 30.167165 30.167248 30.167329 30.167405 30.167479
#> 13 32.274698 32.274738 32.274777 32.274815 32.274852 32.274887 32.274921
#> 14 15.770590 15.770651 15.770707 15.770760 15.770808 15.770852 15.770893
#> 15 18.192933 18.192979 18.193021 18.193059 18.193092 18.193121 18.193146
#> 16 13.049929 13.049823 13.049724 13.049631 13.049543 13.049461 13.049383
#> 17 17.041349 17.041524 17.041688 17.041842 17.041986 17.042121 17.042247
#> 18 27.799458 27.799307 27.799163 27.799024 27.798891 27.798764 27.798642
#> 19 28.992371 28.992535 28.992689 28.992833 28.992968 28.993093 28.993211
#> 20 19.850987 19.851024 19.851062 19.851099 19.851137 19.851175 19.851212
#> 21 23.149869 23.149762 23.149659 23.149560 23.149465 23.149374 23.149286
#>       0.0086    0.0080    0.0075    0.0070   0.0065    0.0061    0.0056
#> 1  20.564509 20.564522 20.564535 20.564548 20.56456 20.564576 20.564590
#> 2  21.129312 21.129377 21.129440 21.129501 21.12956 21.129616 21.129670
#> 3  20.426112 20.426111 20.426112 20.426113 20.42612 20.426120 20.426124
#> 4  15.387611 15.387652 15.387697 15.387745 15.38780 15.387849 15.387904
#> 5  22.602492 22.602371 22.602260 22.602160 22.60207 22.601988 22.601915
#> 6  18.198744 18.198694 18.198643 18.198591 18.19854 18.198485 18.198432
#> 7  20.059254 20.059225 20.059196 20.059165 20.05913 20.059101 20.059067
#> 8  13.661958 13.661822 13.661692 13.661566 13.66144 13.661329 13.661217
#> 9  16.616015 16.615949 16.615886 16.615827 16.61577 16.615719 16.615671
#> 10 11.147906 11.148018 11.148125 11.148226 11.14832 11.148416 11.148505
#> 11  9.615512  9.615545  9.615576  9.615604  9.61563  9.615654  9.615676
#> 12 30.167549 30.167617 30.167681 30.167742 30.16780 30.167856 30.167909
#> 13 32.274954 32.274985 32.275015 32.275045 32.27507 32.275100 32.275126
#> 14 15.770930 15.770964 15.770995 15.771022 15.77105 15.771070 15.771089
#> 15 18.193167 18.193185 18.193200 18.193212 18.19322 18.193228 18.193233
#> 16 13.049311 13.049243 13.049180 13.049121 13.04907 13.049014 13.048967
#> 17 17.042364 17.042474 17.042576 17.042670 17.04276 17.042839 17.042913
#> 18 27.798526 27.798415 27.798309 27.798207 27.79811 27.798019 27.797931
#> 19 28.993320 28.993422 28.993516 28.993604 28.99368 28.993759 28.993827
#> 20 19.851250 19.851287 19.851324 19.851361 19.85140 19.851433 19.851468
#> 21 23.149202 23.149121 23.149043 23.148969 23.14890 23.148830 23.148765
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.4 18.1 14.3 24.4 19.2 17.8 16.4 17.3 10.4 10.4 30.4 33.9 15.5 15.2
#> [16] 13.3 19.2 26.0 30.4 19.7 21.4
#> 
#> $X
#>               am       carb        cyl        disp       drat       gear
#>  [1,]  1.4142136  0.9190183 -0.1714986 -0.63357813  0.6203124  0.5275044
#>  [2,] -0.7071068 -1.3969078 -0.1714986  0.17695447 -0.7944929 -0.8571946
#>  [3,] -0.7071068 -1.3969078 -0.1714986 -0.09597998 -1.3466120 -0.8571946
#>  [4,] -0.7071068  0.9190183  1.0289915  1.02057004 -0.5701945 -0.8571946
#>  [5,] -0.7071068 -0.6249324 -1.3719887 -0.74357899  0.2579842  0.5275044
#>  [6,] -0.7071068  0.9190183 -0.1714986 -0.57072050  0.6548199  0.5275044
#>  [7,] -0.7071068  0.9190183 -0.1714986 -0.57072050  0.6548199  0.5275044
#>  [8,] -0.7071068  0.1470429  1.0289915  0.32417366 -0.8117466 -0.8571946
#>  [9,] -0.7071068  0.1470429  1.0289915  0.32417366 -0.8117466 -0.8571946
#> [10,] -0.7071068  0.9190183  1.0289915  1.94689302 -1.0532987 -0.8571946
#> [11,] -0.7071068  0.9190183  1.0289915  1.84764412 -0.9325227 -0.8571946
#> [12,]  1.4142136 -0.6249324 -1.3719887 -1.33080159  2.3974459  0.5275044
#> [13,]  1.4142136 -1.3969078 -1.3719887 -1.36884700  1.1724316  0.5275044
#> [14,] -0.7071068 -0.6249324  1.0289915  0.67319892 -1.3466120 -0.8571946
#> [15,] -0.7071068 -0.6249324  1.0289915  0.55740855 -0.6737168 -0.8571946
#> [16,] -0.7071068  0.9190183  1.0289915  0.93786263  0.3269991 -0.8571946
#> [17,] -0.7071068 -0.6249324  1.0289915  1.35139967 -0.7944929 -0.8571946
#> [18,]  1.4142136 -0.6249324 -1.3719887 -0.96192655  1.5347598  1.9122034
#> [19,]  1.4142136 -0.6249324 -1.3719887 -1.17034922  0.3960140  1.9122034
#> [20,]  1.4142136  2.4629690 -0.1714986 -0.75763925  0.1372082  1.9122034
#> [21,]  1.4142136 -0.6249324 -1.3719887 -0.95613703  0.9826406  0.5275044
#>               hp        qsec         vs          wt
#>  [1,] -0.5768832 -0.54836419 -0.8660254 -0.41714858
#>  [2,] -0.5768832  1.20986093  1.1547005 -0.07009185
#>  [3,] -0.6668873  1.77656159  1.1547005  0.17999315
#>  [4,]  1.8532267 -1.40568057 -0.8660254  0.29227621
#>  [5,] -1.4409223  1.61672294  1.1547005 -0.09561073
#>  [6,] -0.3428726  0.38160612  1.1547005  0.15957805
#>  [7,] -0.3428726  0.81752971  1.1547005  0.15957805
#>  [8,]  0.6831738 -0.27227925 -0.8660254  0.80265376
#>  [9,]  0.6831738 -0.12697139 -0.8660254  0.45559703
#> [10,]  1.1331941  0.14911354 -0.8660254  2.00714477
#> [11,]  1.3132022  0.03286725 -0.8660254  2.18475616
#> [12,] -1.6209305  0.54144477  1.1547005 -1.70330001
#> [13,] -1.3869199  1.54406901  1.1547005 -1.47873388
#> [14,]  0.1431493 -0.65734509 -0.8660254  0.24123845
#> [15,]  0.1431493 -0.34493319 -0.8660254  0.15447427
#> [16,]  1.8532267 -1.71809247 -0.8660254  0.56788009
#> [17,]  0.5931697 -0.52656801 -0.8660254  0.57298386
#> [18,] -0.9188987 -0.78085677 -0.8660254 -1.16740358
#> [19,] -0.5228808 -0.63554891  1.1547005 -1.80741703
#> [20,]  0.5931697 -1.65270394 -0.8660254 -0.52432787
#> [21,] -0.5948840  0.59956791  1.1547005 -0.51412032
#> attr(,"center")
#>  [1]   0.3333333   2.8095238   6.2857143 236.6047619   3.5404762   3.6190476
#>  [7] 142.0476190  17.7747619   0.4285714   3.2836667
#> attr(,"scale")
#>  [1]   0.4714045   1.2953781   1.6659863 120.9081528   0.5795850   0.7221786
#>  [7]  55.5530430   1.3763880   0.4948717   0.9796669
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 32
#> 
#> $lambda.min
#> [1] 0.6488328
#> 
#> $null.dev
#> [1] 38.28522
#> 
#> $Bias
#> [1] 3.227741
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 11.47403