Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 38.68810 37.38710 33.73368 30.06443 27.08591 24.69381 22.79838 21.32244
#>   [9] 20.19972 19.37337 18.79466 18.42191 18.21949 18.15700 18.20856 18.27513
#>  [17] 17.87155 17.53350 17.47486 17.39864 17.28415 17.14490 17.00812 16.79135
#>  [25] 16.57663 15.54619 15.25516 15.25432 15.05564 14.85665 13.49792 12.45889
#>  [33] 11.87888 12.26883 11.79628 11.33305 10.98896 10.63681 10.63677 11.07059
#>  [41] 11.91655 11.91628 12.40856 13.09859 13.97036 13.73941 13.73881 13.69577
#>  [49] 20.90837 21.93593 20.73712 21.29616 28.85899 28.86178 29.56384 30.52135
#>  [57] 30.52419 36.56517 36.56603 37.57898 37.70056 37.70126 38.63720 38.55571
#>  [65] 38.65995 38.65923 38.65846 38.65762 38.65674 38.65582 38.65485 38.65385
#>  [73] 38.65281 38.65175 38.65052 38.28911 38.35811 38.72978 38.72918 38.72853
#>  [81] 38.72783 38.79660 38.80276 38.80217 38.80152 38.94786 38.94556 38.94328
#>  [89] 38.94100 38.93874 38.93649 38.93426 38.93204 38.92984 38.92767 38.92552
#>  [97] 38.92339 38.92130 38.91923 38.91719
#> 
#> $cvse
#>   [1] 13.052542 13.103371 12.198475 10.980238  9.890281  8.915648  8.048403
#>   [8]  7.284868  6.624868  6.070808  5.626356  5.294612  5.075925  4.966010
#>  [15]  4.955253  4.977738  4.828854  4.700829  4.670315  4.629848  4.551524
#>  [22]  4.456550  4.358710  4.244928  4.136081  3.911927  3.780018  3.621596
#>  [29]  3.423623  3.308750  2.940499  2.790042  2.781905  2.769194  2.740899
#>  [36]  2.525974  2.362568  2.356410  2.356337  2.386628  2.325314  2.325272
#>  [43]  2.561069  2.705864  2.569312  2.576319  2.576191  2.549849  4.598149
#>  [50]  4.650670  4.692198  5.039488  9.906465  9.908146  9.951844  9.941271
#>  [57]  9.942787 12.595984 12.595726 12.492042 12.494969 12.494572 12.449703
#>  [64] 12.461904 12.446744 12.446128 12.445483 12.444811 12.444115 12.443395
#>  [71] 12.442654 12.441893 12.441115 12.440321 12.439490 12.389634 12.395599
#>  [78] 12.386451 12.385611 12.384769 12.383925 12.374757 12.373677 12.372799
#>  [85] 12.371924 12.505927 12.503931 12.501977 12.500066 12.498198 12.496374
#>  [92] 12.494594 12.492857 12.491165 12.489516 12.487911 12.486350 12.484831
#>  [99] 12.483355 12.481922
#> 
#> $fold
#>  [1]  9  1  9 10  1  8  3  8  4  6 10  3  7  1  4  7  5  2  6  2  5
#> 
#> $lambda
#>   [1] 5.245222418 4.891711982 4.562026966 4.254561617 3.967818404 3.700400723
#>   [7] 3.451006099 3.218419838 3.001509113 2.799217444 2.610559557 2.434616580
#>  [13] 2.270531571 2.117505343 1.974792570 1.841698160 1.717573868 1.601815138
#>  [19] 1.493858158 1.393177118 1.299281643 1.211714409 1.130048913 1.053887399
#>  [25] 0.982858915 0.916617514 0.854840561 0.797227169 0.743496726 0.693387535
#>  [31] 0.646655536 0.603073117 0.562428007 0.524522242 0.489171198 0.456202697
#>  [37] 0.425456162 0.396781840 0.370040072 0.345100610 0.321841984 0.300150911
#>  [43] 0.279921744 0.261055956 0.243461658 0.227053157 0.211750534 0.197479257
#>  [49] 0.184169816 0.171757387 0.160181514 0.149385816 0.139317712 0.129928164
#>  [55] 0.121171441 0.113004891 0.105388740 0.098285892 0.091661751 0.085484056
#>  [61] 0.079722717 0.074349673 0.069338753 0.064665554 0.060307312 0.056242801
#>  [67] 0.052452224 0.048917120 0.045620270 0.042545616 0.039678184 0.037004007
#>  [73] 0.034510061 0.032184198 0.030015091 0.027992174 0.026105596 0.024346166
#>  [79] 0.022705316 0.021175053 0.019747926 0.018416982 0.017175739 0.016018151
#>  [85] 0.014938582 0.013931771 0.012992816 0.012117144 0.011300489 0.010538874
#>  [91] 0.009828589 0.009166175 0.008548406 0.007972272 0.007434967 0.006933875
#>  [97] 0.006466555 0.006030731 0.005624280 0.005245222
#> 
#> $fit
#> $beta
#>               5.2452     4.8917    4.5620    4.2546    3.9678    3.7004
#> (Intercept) 19.86667 21.6699042 23.351610 24.919974 26.382635 27.746718
#> am           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000 -0.5519713 -1.066742 -1.546818 -1.994539 -2.412085
#>               3.4510   3.2184    3.0015    2.7992    2.6106    2.4346    2.2705
#> (Intercept) 29.01887 30.20528 31.311726 32.343605 33.305939 34.203415 35.040403
#> am           0.00000  0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.00000  0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.80149 -3.16465 -3.503334 -3.819193 -4.113763 -4.388481 -4.644683
#>                2.1175   1.9748    1.8417    1.7176    1.6018    1.4939
#> (Intercept) 35.820982 36.54895 37.227859 37.703730 37.703730 37.703730
#> am           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> wt          -4.883618 -5.10645 -5.314263 -5.459928 -5.459928 -5.459928
#>                  1.3932     1.2993     1.2117     1.1300     1.0539     0.9829
#> (Intercept) 36.26921420 33.9928345 31.8706763 29.8929277 28.0470907 26.3256558
#> am           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.07388404  0.1911279  0.3004286  0.4022916  0.4973605  0.5860222
#> vs           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.42238379 -5.3628070 -5.3072665 -5.2555054 -5.2071967 -5.1621438
#>                 0.9166     0.8548    0.7972     0.7435    0.6934    0.6467
#> (Intercept) 24.7202398 23.2230235 21.826715 20.5245121 13.646303  7.043831
#> am           0.0000000  0.0000000  0.000000  0.0000000  1.621689  3.187066
#> carb         0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.000000
#> cyl          0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.000000
#> disp         0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.000000
#> drat         0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.000000
#> gear         0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.000000
#> hp           0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.000000
#> qsec         0.6687084  0.7458218  0.817738  0.8848073  1.105874  1.317364
#> vs           0.0000000  0.0000000  0.000000  0.0000000  0.000000  0.000000
#> wt          -5.1201273 -5.0809425 -5.044399 -5.0103178 -4.295506 -3.606481
#>                0.6031    0.5624    0.5245    0.4892    0.4562    0.4255
#> (Intercept)  5.118037  5.115676  5.114249  5.113386  5.112865  5.112549
#> am           3.822101  3.822665  3.823005  3.823211  3.823336  3.823411
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         1.364395  1.364470  1.364516  1.364543  1.364560  1.364570
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -3.346660 -3.346412 -3.346262 -3.346172 -3.346117 -3.346084
#>                0.3968    0.3700    0.3451    0.3218     0.3002     0.2799
#> (Intercept)  5.112359  5.112244  5.112174  5.112132 -30.726459 -30.732741
#> am           3.823457  3.823484  3.823501  3.823511   1.861230   1.860658
#> carb         0.000000  0.000000  0.000000  0.000000   0.000000   0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000   0.000000   0.000000
#> disp         0.000000  0.000000  0.000000  0.000000   0.000000   0.000000
#> drat         0.000000  0.000000  0.000000  0.000000   4.028290   4.029037
#> gear         0.000000  0.000000  0.000000  0.000000   0.000000   0.000000
#> hp           0.000000  0.000000  0.000000  0.000000   0.000000   0.000000
#> qsec         1.364576  1.364580  1.364582  1.364583   2.629321   2.629540
#> vs           0.000000  0.000000  0.000000  0.000000  -4.433047  -4.433666
#> wt          -3.346064 -3.346052 -3.346045 -3.346040  -2.955317  -2.955267
#>                 0.2611     0.2435     0.2271     0.2118     0.1975     0.1842
#> (Intercept) -32.831472 -32.835296 -32.838293 -32.840638 -55.603702 -55.603101
#> am            0.000000   0.000000   0.000000   0.000000   2.411609   2.412182
#> carb         -1.168588  -1.168214  -1.167898  -1.167632  -1.300983  -1.301344
#> cyl           0.000000   0.000000   0.000000   0.000000   1.327110   1.327345
#> disp          0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> drat          6.391678   6.391413   6.391169   6.390947   7.088326   7.088476
#> gear          0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> hp            0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> qsec          2.288736   2.289051   2.289308   2.289519   2.934875   2.934690
#> vs           -4.694393  -4.695064  -4.695612  -4.696061  -3.914096  -3.913366
#> wt           -1.904151  -1.904613  -1.905011  -1.905353  -2.075285  -2.074955
#>                 0.1718     0.1602     0.1494     0.1393     0.1299     0.1212
#> (Intercept) -55.602150 -55.600959 -55.599615 -55.598186 -55.596726 -55.595276
#> am            2.412634   2.412985   2.413252   2.413450   2.413593   2.413690
#> carb         -1.301655  -1.301924  -1.302156  -1.302354  -1.302525  -1.302670
#> cyl           1.327520   1.327646   1.327733   1.327789   1.327820   1.327833
#> disp          0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> drat          7.088600   7.088701   7.088785   7.088855   7.088913   7.088962
#> gear          0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> hp            0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> qsec          2.934516   2.934353   2.934203   2.934065   2.933938   2.933824
#> vs           -3.912747  -3.912225  -3.911785  -3.911417  -3.911110  -3.910855
#> wt           -2.074656  -2.074387  -2.074144  -2.073926  -2.073731  -2.073558
#>                 0.1130     0.1054     0.0983     0.0917     0.0855     0.0797
#> (Intercept) -55.593864 -55.592514 -55.591240 -55.590051 -55.588952 -55.587946
#> am            2.413751   2.413784   2.413794   2.413787   2.413768   2.413739
#> carb         -1.302794  -1.302900  -1.302989  -1.303065  -1.303128  -1.303182
#> cyl           1.327831   1.327818   1.327798   1.327772   1.327743   1.327712
#> disp          0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> drat          7.089003   7.089038   7.089068   7.089094   7.089117   7.089137
#> gear          0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> hp            0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> qsec          2.933721   2.933628   2.933545   2.933471   2.933405   2.933347
#> vs           -3.910645  -3.910472  -3.910331  -3.910217  -3.910125  -3.910052
#> wt           -2.073403  -2.073267  -2.073146  -2.073040  -2.072947  -2.072865
#>                 0.0743     0.0693     0.0647       0.0603       0.0562
#> (Intercept) -55.587032 -55.586207 -55.585468 -48.41720270 -48.40793405
#> am            2.413704   2.413665   2.413623   2.37777328   2.37776432
#> carb         -1.303227  -1.303264  -1.303296  -1.33333092  -1.33330550
#> cyl           1.327680   1.327649   1.327618   1.44288889   1.44281811
#> disp          0.000000   0.000000   0.000000   0.00000000   0.00000000
#> drat          7.089155   7.089171   7.089185   6.95507062   6.95456176
#> gear          0.000000   0.000000   0.000000   0.00000000   0.00000000
#> hp            0.000000   0.000000   0.000000  -0.01180212  -0.01180929
#> qsec          2.933297   2.933252   2.933213   2.54558573   2.54520335
#> vs           -3.909994  -3.909950  -3.909916  -3.35828728  -3.35774722
#> wt           -2.072794  -2.072732  -2.072678  -1.73339638  -1.73320856
#>                   0.0525       0.0489       0.0456       0.0425       0.0397
#> (Intercept) -48.39933527 -48.39135959 -48.38396310 -48.37710468 -48.37074581
#> am            2.37775304   2.37774049   2.37772748   2.37771465   2.37770251
#> carb         -1.33328112  -1.33325784  -1.33323566  -1.33321460  -1.33319467
#> cyl           1.44275348   1.44269475   1.44264161   1.44259373   1.44255076
#> disp          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> drat          6.95409156   6.95365688   6.95325482   6.95288274   6.95253822
#> gear          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> hp           -0.01181599  -0.01182226  -0.01182811  -0.01183358  -0.01183868
#> qsec          2.54484816   2.54451829   2.54421196   2.54392751   2.54366342
#> vs           -3.35724526  -3.35677829  -3.35634352  -3.35593844  -3.35556079
#> wt           -1.73303408  -1.73287214  -1.73272196  -1.73258278  -1.73245390
#>                   0.0370       0.0345       0.0322      0.0300        0.0280
#> (Intercept) -48.36485051 -48.35938520 -48.35431857 -48.3496215 -46.501408102
#> am            2.37769139   2.37768155   2.37767314   2.3776662   2.559865971
#> carb         -1.33317584  -1.33315809  -1.33314141  -1.3331257  -1.165381042
#> cyl           1.44251233   1.44247808   1.44244765   1.4424207   1.333406458
#> disp          0.00000000   0.00000000   0.00000000   0.0000000   0.004113588
#> drat          6.95221904   6.95192316   6.95164873   6.9513941   6.808177968
#> gear          0.00000000   0.00000000   0.00000000   0.0000000  -0.205878802
#> hp           -0.01184345  -0.01184789  -0.01185204  -0.0118559  -0.014562910
#> qsec          2.54341823   2.54319060   2.54297930   2.5427831   2.548889408
#> vs           -3.35520850  -3.35487974  -3.35457281  -3.3542862  -3.244308220
#> wt           -1.73233463  -1.73222431  -1.73212233  -1.7320281  -2.086419774
#>                    0.0261        0.0243        0.0227        0.0212
#> (Intercept) -46.501844460 -46.502463667 -46.503248178 -46.504181365
#> am            2.559731298   2.559592369   2.559450043   2.559305116
#> carb         -1.165308243  -1.165249977  -1.165205103  -1.165172539
#> cyl           1.333116287   1.332851691   1.332611269   1.332393663
#> disp          0.004112194   0.004110714   0.004109161   0.004107547
#> drat          6.808024995   6.807907761   6.807823425   6.807769283
#> gear         -0.205919901  -0.205952039  -0.205975878  -0.205992054
#> hp           -0.014555499  -0.014548458  -0.014541776  -0.014535443
#> qsec          2.549034626   2.549173189   2.549305281   2.549431088
#> vs           -3.244943687  -3.245549456  -3.246126172  -3.246674516
#> wt           -2.086498133  -2.086551667  -2.086582537  -2.086592784
#>                    0.0197        0.0184        0.0172        0.0160
#> (Intercept) -46.505247499 -46.506431729 -46.507720061 -46.509099335
#> am            2.559158323   2.559010340   2.558861785   2.558713227
#> carb         -1.165151258  -1.165140287  -1.165138705  -1.165145643
#> cyl           1.332197558   1.332021678   1.331864795   1.331725723
#> disp          0.004105885   0.004104185   0.004102457   0.004100709
#> drat          6.807742769   6.807741451   6.807763030   6.807805331
#> gear         -0.206001172  -0.206003812  -0.206000527  -0.205991840
#> hp           -0.014529448  -0.014523782  -0.014518433  -0.014513391
#> qsec          2.549550800   2.549664606   2.549772694   2.549875254
#> vs           -3.247195207  -3.247688990  -3.248156634  -3.248598929
#> wt           -2.086584332  -2.086558993  -2.086518469  -2.086464360
#>                   0.0149        0.0139        0.0130        0.0121
#> (Intercept) -46.51055720 -46.512082088 -46.513663195 -46.515290443
#> am            2.55856518   2.558418115   2.558272454   2.558128578
#> carb         -1.16516028  -1.165181849  -1.165209618  -1.165242910
#> cyl           1.33160332   1.331496494   1.331404189   1.331325397
#> disp          0.00409895   0.004097188   0.004095429   0.004093679
#> drat          6.80786630   6.807944019   6.808036660   6.808142526
#> gear         -0.20597825  -0.205960232  -0.205938229  -0.205912665
#> hp           -0.01450864  -0.014504182  -0.014499994  -0.014496069
#> qsec          2.54997247   2.550064541   2.550151641   2.550233959
#> vs           -3.24901668  -3.249410695  -3.249781801  -3.250130822
#> wt           -2.08639816  -2.086321277  -2.086235013  -2.086140589
#>                    0.0113        0.0105        0.0098       0.0092
#> (Intercept) -46.516954462 -46.518646561 -46.520358697 -46.52208345
#> am            2.557986828   2.557847505   2.557710877   2.55757718
#> carb         -1.165281086  -1.165323549  -1.165369744  -1.16541915
#> cyl           1.331259155   1.331204543   1.331160682   1.33112674
#> disp          0.004091945   0.004090231   0.004088541   0.00408688
#> drat          6.808260021   6.808387654   6.808524032   6.80866786
#> gear         -0.205883937  -0.205852420  -0.205818465  -0.20578240
#> hp           -0.014492396  -0.014488965  -0.014485765  -0.01448279
#> qsec          2.550311675   2.550384970   2.550454019   2.55051900
#> vs           -3.250458582  -3.250765906  -3.251053612  -3.25132251
#> wt           -2.086039140  -2.085931720  -2.085819303  -2.08570279
#>                    0.0085        0.0080       0.0074        0.0069
#> (Intercept) -46.523813995 -46.525544079 -46.52726799 -46.528980513
#> am            2.557446609   2.557319345   2.55719553   2.557075288
#> carb         -1.165471286  -1.165525704  -1.16558199  -1.165639762
#> cyl           1.331101908   1.331085445   1.33107663   1.331074788
#> disp          0.004085251   0.004083657   0.00408210   0.004080584
#> drat          6.808817935   6.808973139   6.80913244   6.809294881
#> gear         -0.205744535  -0.205705156  -0.20566453  -0.205622903
#> hp           -0.014480017  -0.014477449  -0.01447507  -0.014472876
#> qsec          2.550580074   2.550637416   2.55069119   2.550741549
#> vs           -3.251573407  -3.251807090  -3.25202434  -3.252225916
#> wt           -2.085583021  -2.085460753  -2.08533669  -2.085211481
#>                    0.0065        0.0060        0.0056       0.0052
#> (Intercept) -46.530676952 -46.532353051 -46.534004999 -46.53562940
#> am            2.556958714   2.556845884   2.556736854   2.55663166
#> carb         -1.165698666  -1.165758379  -1.165818604  -1.16587907
#> cyl           1.331079275   1.331089490   1.331104861   1.33112485
#> disp          0.004079109   0.004077678   0.004076291   0.00407495
#> drat          6.809459589   6.809625757   6.809792649   6.80995959
#> gear         -0.205580509  -0.205537558  -0.205494247  -0.20545075
#> hp           -0.014470853  -0.014468994  -0.014467289  -0.01446573
#> qsec          2.550788654   2.550832655   2.550873700   2.55091193
#> vs           -3.252412572  -3.252585037  -3.252744026  -3.25289023
#> wt           -2.085085704  -2.084959897  -2.084834540  -2.08471007
#> 
#> $iter
#>   [1]  0  4  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  5  5  5  4  4  4
#>  [26]  4  4  4  4 50 46 14  1  1  1  1  1  1  1  1  1 71  1 69  1  1  1 68  1  1
#>  [51]  1  1  1  1  1  1  1  1  1  1  1  1  1  1 68  1  1  1  1  1  1  1  1  1  1
#>  [76] 88  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> 
#> $lambda
#>   [1] 5.245222418 4.891711982 4.562026966 4.254561617 3.967818404 3.700400723
#>   [7] 3.451006099 3.218419838 3.001509113 2.799217444 2.610559557 2.434616580
#>  [13] 2.270531571 2.117505343 1.974792570 1.841698160 1.717573868 1.601815138
#>  [19] 1.493858158 1.393177118 1.299281643 1.211714409 1.130048913 1.053887399
#>  [25] 0.982858915 0.916617514 0.854840561 0.797227169 0.743496726 0.693387535
#>  [31] 0.646655536 0.603073117 0.562428007 0.524522242 0.489171198 0.456202697
#>  [37] 0.425456162 0.396781840 0.370040072 0.345100610 0.321841984 0.300150911
#>  [43] 0.279921744 0.261055956 0.243461658 0.227053157 0.211750534 0.197479257
#>  [49] 0.184169816 0.171757387 0.160181514 0.149385816 0.139317712 0.129928164
#>  [55] 0.121171441 0.113004891 0.105388740 0.098285892 0.091661751 0.085484056
#>  [61] 0.079722717 0.074349673 0.069338753 0.064665554 0.060307312 0.056242801
#>  [67] 0.052452224 0.048917120 0.045620270 0.042545616 0.039678184 0.037004007
#>  [73] 0.034510061 0.032184198 0.030015091 0.027992174 0.026105596 0.024346166
#>  [79] 0.022705316 0.021175053 0.019747926 0.018416982 0.017175739 0.016018151
#>  [85] 0.014938582 0.013931771 0.012992816 0.012117144 0.011300489 0.010538874
#>  [91] 0.009828589 0.009166175 0.008548406 0.007972272 0.007434967 0.006933875
#>  [97] 0.006466555 0.006030731 0.005624280 0.005245222
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 29
#> 
#> $loss
#>   [1] 791.98667 681.07431 588.27963 510.99535 446.97048 394.26276 351.19743
#>   [8] 316.33139 288.42211 266.40067 249.34836 236.47636 227.10814 220.66416
#>  [15] 216.64872 214.63837 214.22714 214.22714 214.22714 207.91180 198.76545
#>  [22] 191.20589 185.00088 179.94098 175.85835 172.60426 170.05078 168.08801
#>  [29] 166.62162 144.02384 133.95145 133.08516 133.08516 133.08516 133.08516
#>  [36] 133.08516 133.08516 133.08516 133.08516 133.08516 133.08516 106.39580
#>  [43] 106.39579 102.41891 102.41889 102.41888 102.41887  89.54805  89.54804
#>  [50]  89.54802  89.54802  89.54801  89.54801  89.54800  89.54800  89.54800
#>  [57]  89.54800  89.54800  89.54800  89.54800  89.54800  89.54800  89.54800
#>  [64]  89.54800  88.56421  88.56420  88.56419  88.56417  88.56417  88.56416
#>  [71]  88.56415  88.56414  88.56414  88.56413  88.56413  88.17527  88.17526
#>  [78]  88.17524  88.17523  88.17522  88.17521  88.17521  88.17520  88.17519
#>  [85]  88.17519  88.17518  88.17518  88.17517  88.17517  88.17517  88.17516
#>  [92]  88.17516  88.17516  88.17516  88.17515  88.17515  88.17515  88.17515
#>  [99]  88.17515  88.17514
#> 
#> $linear.predictors
#>      5.2452   4.8917   4.5620   4.2546   3.9678   3.7004   3.4510   3.2184
#> 1  19.86667 20.08299 20.28473 20.47287 20.64834 20.81197 20.96458 21.10691
#> 2  19.86667 20.38933 20.87677 21.33136 21.75531 22.15068 22.51941 22.86329
#> 3  19.86667 19.77112 19.68202 19.59892 19.52142 19.44915 19.38174 19.31888
#> 4  19.86667 19.76008 19.66068 19.56798 19.48153 19.40090 19.32571 19.25559
#> 5  19.86667 19.69937 19.54334 19.39783 19.26213 19.13558 19.01755 18.90748
#> 6  19.86667 19.77112 19.68202 19.59892 19.52142 19.44915 19.38174 19.31888
#> 7  19.86667 19.42338 19.00997 18.62442 18.26486 17.92953 17.61680 17.32515
#> 8  19.86667 19.61105 19.37266 19.15034 18.94301 18.74964 18.56931 18.40113
#> 9  19.86667 19.58345 19.31933 19.07300 18.84328 18.62904 18.42924 18.24290
#> 10 19.86667 18.77206 17.75122 16.79918 15.91131 15.08327 14.31105 13.59086
#> 11 19.86667 18.71962 17.64988 16.65223 15.72182 14.85412 14.04490 13.29022
#> 12 19.86667 20.45557 21.00478 21.51697 21.99465 22.44013 22.85559 23.24305
#> 13 19.86667 20.77847 21.62882 22.42186 23.16145 23.85120 24.49446 25.09437
#> 14 19.86667 20.65704 21.39414 22.08156 22.72266 23.32054 23.87813 24.39814
#> 15 19.86667 20.30930 20.72209 21.10707 21.46610 21.80093 22.11319 22.40441
#> 16 19.86667 19.77388 19.68735 19.60665 19.53139 19.46121 19.39575 19.33470
#> 17 19.86667 19.55033 19.25532 18.98019 18.72361 18.48431 18.26115 18.05302
#> 18 19.86667 19.54757 19.24999 18.97246 18.71363 18.47225 18.24714 18.03720
#> 19 19.86667 20.48869 21.06878 21.60978 22.11432 22.58486 23.02368 23.43293
#> 20 19.86667 19.92016 19.97004 20.01656 20.05995 20.10041 20.13814 20.17334
#> 21 19.86667 20.13542 20.38607 20.61982 20.83782 21.04112 21.23073 21.40755
#>      3.0015   2.7992   2.6106   2.4346   2.2705    2.1175    1.9748    1.8417
#> 1  21.23964 21.36343 21.47887 21.58653 21.68694 21.780580 21.867909 21.949352
#> 2  23.18399 23.48308 23.76201 24.02214 24.26474 24.490988 24.701988 24.898768
#> 3  19.26026 19.20558 19.15459 19.10704 19.06269 19.021336 18.982765 18.946793
#> 4  19.19019 19.12920 19.07232 19.01927 18.96980 18.923663 18.880636 18.840508
#> 5  18.80482 18.70909 18.61980 18.53654 18.45889 18.386465 18.318926 18.255939
#> 6  19.26026 19.20558 19.15459 19.10704 19.06269 19.021336 18.982765 18.946793
#> 7  17.05316 16.79949 16.56292 16.34230 16.13654 15.944656 15.765701 15.598807
#> 8  18.24429 18.09802 17.96160 17.83438 17.71574 17.605086 17.501894 17.405657
#> 9  18.06912 17.90706 17.75591 17.61496 17.48350 17.360905 17.246572 17.139943
#> 10 12.91922 12.29284 11.70868 11.16389 10.65582 10.181987  9.740090  9.327976
#> 11 12.58640 11.93002 11.31788 10.74699 10.21457  9.718043  9.254978  8.823121
#> 12 23.60439 23.94138 24.25566 24.54876 24.82210 25.077022 25.314762 25.536480
#> 13 25.65384 26.17561 26.66221 27.11602 27.53924 27.933939 28.302035 28.645324
#> 14 24.88311 25.33539 25.75718 26.15055 26.51741 26.859543 27.178616 27.476186
#> 15 22.67601 22.92930 23.16551 23.38581 23.59126 23.782863 23.961553 24.128200
#> 16 19.27777 19.22468 19.17516 19.12898 19.08592 19.045754 19.008297 18.973364
#> 17 17.85892 17.67791 17.50909 17.35165 17.20482 17.067888 16.940185 16.821088
#> 18 17.84141 17.65881 17.48852 17.32971 17.18160 17.043470 16.914652 16.794516
#> 19 23.81459 24.17053 24.50249 24.81207 25.10078 25.370039 25.621149 25.855335
#> 20 20.20616 20.23676 20.26531 20.29193 20.31676 20.339913 20.361506 20.381644
#> 21 21.57246 21.72625 21.86968 22.00344 22.12818 22.244524 22.353021 22.454207
#>       1.7176    1.6018    1.4939    1.3932    1.2993    1.2117    1.1300
#> 1  22.006438 22.006438 22.006438 21.937367 21.827761 21.725580 21.630353
#> 2  25.036698 25.036698 25.036698 25.064266 25.108012 25.148794 25.186802
#> 3  18.921579 18.921579 18.921579 18.873720 18.797775 18.726974 18.660992
#> 4  18.812381 18.812381 18.812381 19.001702 19.302128 19.582201 19.843215
#> 5  18.211789 18.211789 18.211789 18.081627 17.875079 17.682524 17.503072
#> 6  18.921579 18.921579 18.921579 18.968292 19.042418 19.111523 19.175925
#> 7  15.481825 15.481825 15.481825 15.485694 15.491835 15.497559 15.502894
#> 8  17.338200 17.338200 17.338200 17.344082 17.353415 17.362116 17.370225
#> 9  17.065204 17.065204 17.065204 17.102516 17.161726 17.216924 17.268366
#> 10  9.039110  9.039110  9.039110  9.130134  9.274577  9.409234  9.534727
#> 11  8.520417  8.520417  8.520417  8.573633  8.658079  8.736803  8.810171
#> 12 25.691889 25.691889 25.691889 25.778492 25.915919 26.044035 26.163433
#> 13 28.885947 28.885947 28.885947 28.880397 28.871589 28.863379 28.855727
#> 14 27.684763 27.684763 27.684763 27.789432 27.955528 28.110372 28.254678
#> 15 24.245009 24.245009 24.245009 24.381458 24.597984 24.799841 24.987962
#> 16 18.948879 18.948879 18.948879 18.921520 18.878105 18.837631 18.799911
#> 17 16.737608 16.737608 16.737608 16.585813 16.344936 16.120378 15.911100
#> 18 16.710309 16.710309 16.710309 16.679871 16.631572 16.586544 16.544581
#> 19 26.019485 26.019485 26.019485 25.899176 25.708263 25.530284 25.364416
#> 20 20.395760 20.395760 20.395760 20.151576 19.764090 19.402856 19.066204
#> 21 22.525131 22.525131 22.525131 22.569230 22.639209 22.704448 22.765246
#>       1.0539    0.9829    0.9166    0.8548    0.7972    0.7435    0.6934
#> 1  21.541477 21.458590 21.381291 21.309200 21.241969 21.179269 21.740391
#> 2  25.222274 25.255356 25.286208 25.314980 25.341814 25.366839 25.882737
#> 3  18.599410 18.541979 18.488419 18.438468 18.391884 18.348439 17.691741
#> 4  20.086820 20.314007 20.525883 20.723479 20.907757 21.079616 21.144629
#> 5  17.335589 17.179394 17.033726 16.897876 16.771181 16.653025 15.828394
#> 6  19.236032 19.292088 19.344365 19.393120 19.438588 19.480992 19.107260
#> 7  15.507874 15.512517 15.516848 15.520886 15.524653 15.528166 15.405805
#> 8  17.377793 17.384850 17.391433 17.397571 17.403296 17.408635 17.087452
#> 9  17.316377 17.361152 17.402910 17.441853 17.478171 17.512042 17.315026
#> 10  9.651851  9.761080  9.862948  9.957951 10.046551 10.129179 10.978514
#> 11  8.878645  8.942504  9.002060  9.057601  9.109399  9.157707  9.951152
#> 12 26.274868 26.378792 26.475712 26.566100 26.650396 26.729011 27.349250
#> 13 28.848585 28.841925 28.835714 28.829921 28.824518 28.819480 28.811540
#> 14 28.389359 28.514964 28.632103 28.741347 28.843229 28.938244 29.392635
#> 15 25.163535 25.327276 25.479981 25.622394 25.755209 25.879073 25.186424
#> 16 18.764707 18.731876 18.701258 18.672703 18.646072 18.621237 18.022864
#> 17 15.715781 15.533626 15.363747 15.205318 15.057566 14.919772 14.193081
#> 18 16.505417 16.468892 16.434828 16.403061 16.373434 16.345805 15.985238
#> 19 25.209611 25.065239 24.930597 24.805030 24.687926 24.578714 24.543708
#> 20 18.752005 18.458982 18.185708 17.930852 17.693172 17.471511 17.686414
#> 21 22.821990 22.874909 22.924262 22.970288 23.013213 23.053244 23.895746
#>      0.6467   0.6031   0.5624   0.5245   0.4892   0.4562   0.4255   0.3968
#> 1  22.28381 22.54050 22.54069 22.54081 22.54088 22.54093 22.54095 22.54097
#> 2  26.38001 26.56728 26.56746 26.56757 26.56763 26.56767 26.56770 26.56771
#> 3  17.05908 16.82753 16.82731 16.82717 16.82708 16.82703 16.82700 16.82699
#> 4  21.20251 21.12666 21.12668 21.12669 21.12670 21.12670 21.12671 21.12671
#> 5  15.03575 14.78248 14.78220 14.78202 14.78192 14.78186 14.78182 14.78180
#> 6  18.74530 18.57396 18.57383 18.57375 18.57370 18.57367 18.57365 18.57364
#> 7  15.28759 15.23761 15.23756 15.23754 15.23752 15.23751 15.23751 15.23750
#> 8  16.77727 16.64835 16.64824 16.64817 16.64813 16.64811 16.64809 16.64808
#> 9  17.12389 17.02678 17.02671 17.02666 17.02664 17.02662 17.02661 17.02661
#> 10 11.79602 12.07990 12.08019 12.08037 12.08048 12.08054 12.08058 12.08060
#> 11 10.71568 10.99790 10.99818 10.99834 10.99845 10.99851 10.99854 10.99856
#> 12 27.94572 28.14226 28.14247 28.14260 28.14268 28.14273 28.14276 28.14277
#> 13 28.80402 28.80388 28.80388 28.80388 28.80387 28.80387 28.80387 28.80387
#> 14 29.82856 29.95048 29.95064 29.95073 29.95079 29.95082 29.95084 29.95085
#> 15 24.51432 24.17007 24.16982 24.16968 24.16959 24.16953 24.16950 24.16948
#> 16 17.44597 17.22630 17.22609 17.22596 17.22589 17.22584 17.22581 17.22580
#> 17 13.49553 13.29219 13.29194 13.29179 13.29170 13.29165 13.29161 13.29159
#> 18 15.63797 15.51307 15.51294 15.51287 15.51282 15.51279 15.51278 15.51277
#> 19 24.51301 24.56368 24.56367 24.56367 24.56367 24.56366 24.56366 24.56366
#> 20 17.90014 18.11496 18.11504 18.11508 18.11511 18.11513 18.11514 18.11515
#> 21 24.70786 25.01417 25.01446 25.01464 25.01475 25.01481 25.01485 25.01488
#>      0.3700   0.3451   0.3218   0.3002   0.2799   0.2611   0.2435   0.2271
#> 1  22.54098 22.54098 22.54099 23.09961 23.09953 20.90157 20.90223 20.90279
#> 2  26.56772 26.56772 26.56773 24.28597 24.28556 24.08925 24.08889 24.08860
#> 3  16.82697 16.82697 16.82696 16.54741 16.54737 17.36914 17.36900 17.36888
#> 4  21.12671 21.12671 21.12671 18.89805 18.89780 18.63646 18.63637 18.63630
#> 5  14.78178 14.78178 14.78177 13.30232 13.30207 12.46722 12.46737 12.46752
#> 6  18.57364 18.57363 18.57363 18.58168 18.58187 18.18875 18.18888 18.18898
#> 7  15.23750 15.23750 15.23750 15.36244 15.36245 15.35933 15.35940 15.35947
#> 8  16.64808 16.64807 16.64807 16.89311 16.89315 16.46448 16.46478 16.46503
#> 9  17.02661 17.02660 17.02660 17.79708 17.79721 17.28477 17.28517 17.28551
#> 10 12.08062 12.08063 12.08063 12.83621 12.83631 12.37647 12.37660 12.37669
#> 11 10.99858 10.99859 10.99859 12.29153 12.29173 12.83139 12.83121 12.83105
#> 12 28.14278 28.14279 28.14279 27.82834 27.82827 27.75615 27.75606 27.75597
#> 13 28.80387 28.80387 28.80387 30.48339 30.48372 30.96012 30.96014 30.96015
#> 14 29.95086 29.95087 29.95087 30.60159 30.60171 30.33016 30.33033 30.33047
#> 15 24.16947 24.16946 24.16946 25.07303 25.07338 26.05863 26.05868 26.05872
#> 16 17.22579 17.22578 17.22578 17.29840 17.29842 18.01951 18.01945 18.01941
#> 17 13.29158 13.29157 13.29157 13.46849 13.46855 14.29262 14.29237 14.29217
#> 18 15.51276 15.51276 15.51276 15.14741 15.14734 16.21921 16.21890 16.21865
#> 19 24.56366 24.56366 24.56366 26.56538 26.56559 27.29349 27.29351 27.29351
#> 20 18.11515 18.11515 18.11516 16.89096 16.89058 16.61757 16.61722 16.61693
#> 21 25.01489 25.01490 25.01490 23.94759 23.94739 23.68371 23.68343 23.68320
#>      0.2118   0.1975   0.1842   0.1718   0.1602   0.1494   0.1393   0.1299
#> 1  20.90326 21.19624 21.19576 21.19534 21.19497 21.19464 21.19436 21.19411
#> 2  24.08835 23.99469 23.99507 23.99540 23.99570 23.99595 23.99617 23.99637
#> 3  17.36879 17.55203 17.55225 17.55243 17.55257 17.55269 17.55278 17.55286
#> 4  18.63624 18.87035 18.87054 18.87070 18.87083 18.87093 18.87102 18.87109
#> 5  12.46765 11.64243 11.64219 11.64200 11.64184 11.64171 11.64160 11.64151
#> 6  18.18907 17.59640 17.59604 17.59573 17.59548 17.59527 17.59510 17.59496
#> 7  15.35952 15.49181 15.49179 15.49176 15.49174 15.49172 15.49169 15.49167
#> 8  16.46525 16.78438 16.78421 16.78405 16.78390 16.78377 16.78364 16.78352
#> 9  17.28579 17.85457 17.85434 17.85412 17.85392 17.85374 17.85357 17.85341
#> 10 12.37676 12.45185 12.45173 12.45163 12.45155 12.45148 12.45142 12.45138
#> 11 12.83091 12.73767 12.73773 12.73779 12.73785 12.73792 12.73798 12.73804
#> 12 27.75590 28.39803 28.39825 28.39842 28.39857 28.39868 28.39878 28.39885
#> 13 30.96016 31.54804 31.54800 31.54796 31.54792 31.54788 31.54783 31.54779
#> 14 30.33058 31.40987 31.40991 31.40992 31.40991 31.40988 31.40985 31.40980
#> 15 26.05876 24.32774 24.32731 24.32698 24.32671 24.32651 24.32637 24.32626
#> 16 18.01938 18.38418 18.38434 18.38446 18.38456 18.38464 18.38469 18.38473
#> 17 14.29200 13.50603 13.50604 13.50607 13.50611 13.50615 13.50620 13.50625
#> 18 16.21844 16.30341 16.30374 16.30402 16.30427 16.30447 16.30465 16.30479
#> 19 27.29351 25.48697 25.48664 25.48640 25.48622 25.48609 25.48602 25.48597
#> 20 16.61669 18.11063 18.11123 18.11173 18.11213 18.11246 18.11272 18.11292
#> 21 23.68301 23.55269 23.55290 23.55310 23.55327 23.55343 23.55357 23.55370
#>      0.1212   0.1130   0.1054   0.0983   0.0917   0.0855   0.0797   0.0743
#> 1  21.19389 21.19370 21.19353 21.19338 21.19326 21.19315 21.19306 21.19297
#> 2  23.99653 23.99668 23.99680 23.99690 23.99699 23.99707 23.99713 23.99719
#> 3  17.55292 17.55297 17.55301 17.55303 17.55306 17.55307 17.55308 17.55309
#> 4  18.87114 18.87118 18.87122 18.87124 18.87126 18.87127 18.87128 18.87129
#> 5  11.64144 11.64139 11.64134 11.64131 11.64128 11.64126 11.64124 11.64123
#> 6  17.59486 17.59477 17.59470 17.59465 17.59461 17.59458 17.59456 17.59455
#> 7  15.49165 15.49162 15.49160 15.49159 15.49157 15.49155 15.49154 15.49152
#> 8  16.78342 16.78333 16.78324 16.78316 16.78310 16.78303 16.78298 16.78293
#> 9  17.85327 17.85314 17.85303 17.85292 17.85283 17.85275 17.85268 17.85261
#> 10 12.45134 12.45131 12.45129 12.45127 12.45126 12.45124 12.45124 12.45123
#> 11 12.73810 12.73816 12.73821 12.73826 12.73830 12.73834 12.73838 12.73842
#> 12 28.39891 28.39896 28.39899 28.39902 28.39903 28.39905 28.39906 28.39906
#> 13 31.54776 31.54772 31.54769 31.54766 31.54763 31.54761 31.54759 31.54757
#> 14 31.40976 31.40971 31.40966 31.40961 31.40956 31.40951 31.40947 31.40943
#> 15 24.32619 24.32614 24.32611 24.32611 24.32611 24.32612 24.32614 24.32617
#> 16 18.38476 18.38478 18.38479 18.38479 18.38479 18.38479 18.38478 18.38478
#> 17 13.50630 13.50635 13.50640 13.50645 13.50650 13.50654 13.50658 13.50662
#> 18 16.30492 16.30502 16.30511 16.30518 16.30524 16.30529 16.30533 16.30537
#> 19 25.48595 25.48596 25.48597 25.48600 25.48604 25.48608 25.48612 25.48617
#> 20 18.11308 18.11321 18.11330 18.11337 18.11342 18.11345 18.11347 18.11348
#> 21 23.55382 23.55392 23.55401 23.55409 23.55416 23.55422 23.55427 23.55432
#>      0.0693   0.0647   0.0603   0.0562   0.0525   0.0489   0.0456   0.0425
#> 1  21.19291 21.19285 21.45348 21.45367 21.45385 21.45403 21.45418 21.45433
#> 2  23.99723 23.99727 24.07180 24.07204 24.07226 24.07245 24.07264 24.07281
#> 3  17.55309 17.55310 17.66533 17.66537 17.66540 17.66543 17.66546 17.66548
#> 4  18.87129 18.87129 18.97948 18.97965 18.97980 18.97995 18.98009 18.98021
#> 5  11.64122 11.64122 11.36069 11.36072 11.36075 11.36077 11.36079 11.36081
#> 6  17.59454 17.59454 17.98207 17.98233 17.98257 17.98280 17.98302 17.98322
#> 7  15.49151 15.49150 15.59187 15.59191 15.59194 15.59197 15.59200 15.59203
#> 8  16.78289 16.78285 16.69034 16.69024 16.69014 16.69005 16.68997 16.68990
#> 9  17.85256 17.85251 17.62190 17.62166 17.62143 17.62122 17.62102 17.62084
#> 10 12.45123 12.45122 12.42081 12.42076 12.42072 12.42068 12.42064 12.42061
#> 11 12.73845 12.73847 12.62207 12.62193 12.62179 12.62167 12.62155 12.62144
#> 12 28.39906 28.39906 28.38734 28.38730 28.38726 28.38723 28.38719 28.38716
#> 13 31.54755 31.54754 31.72678 31.72668 31.72660 31.72653 31.72646 31.72640
#> 14 31.40939 31.40936 31.10014 31.09981 31.09949 31.09920 31.09893 31.09868
#> 15 24.32620 24.32623 23.91604 23.91582 23.91562 23.91544 23.91526 23.91510
#> 16 18.38477 18.38476 18.68182 18.68192 18.68202 18.68211 18.68220 18.68228
#> 17 13.50666 13.50669 13.41471 13.41469 13.41467 13.41465 13.41463 13.41462
#> 18 16.30539 16.30542 16.55282 16.55295 16.55308 16.55319 16.55330 16.55340
#> 19 25.48621 25.48626 25.60425 25.60438 25.60451 25.60463 25.60473 25.60483
#> 20 18.11349 18.11348 17.82112 17.82093 17.82075 17.82057 17.82041 17.82026
#> 21 23.55436 23.55439 23.53514 23.53524 23.53534 23.53543 23.53550 23.53558
#>      0.0397   0.0370   0.0345   0.0322   0.0300   0.0280   0.0261   0.0243
#> 1  21.45447 21.45460 21.45472 21.45484 21.45494 21.56565 21.56570 21.56575
#> 2  24.07296 24.07311 24.07324 24.07336 24.07347 24.05462 24.05463 24.05464
#> 3  17.66551 17.66553 17.66555 17.66557 17.66559 17.80040 17.80018 17.79997
#> 4  18.98033 18.98044 18.98054 18.98064 18.98073 18.97826 18.97810 18.97795
#> 5  11.36082 11.36083 11.36085 11.36086 11.36087 11.57980 11.58006 11.58028
#> 6  17.98341 17.98359 17.98376 17.98391 17.98406 17.82333 17.82311 17.82291
#> 7  15.59206 15.59209 15.59211 15.59214 15.59216 15.32532 15.32535 15.32538
#> 8  16.68983 16.68977 16.68971 16.68965 16.68960 16.54448 16.54456 16.54464
#> 9  17.62067 17.62052 17.62037 17.62024 17.62012 17.45972 17.45985 17.45999
#> 10 12.42057 12.42054 12.42051 12.42049 12.42046 12.66619 12.66621 12.66622
#> 11 12.62133 12.62124 12.62115 12.62106 12.62098 12.58735 12.58747 12.58758
#> 12 28.38714 28.38711 28.38708 28.38706 28.38704 28.33559 28.33554 28.33549
#> 13 31.72634 31.72630 31.72625 31.72621 31.72617 31.94781 31.94751 31.94723
#> 14 31.09845 31.09823 31.09804 31.09785 31.09768 31.12960 31.12962 31.12964
#> 15 23.91495 23.91480 23.91467 23.91454 23.91443 23.93684 23.93726 23.93765
#> 16 18.68236 18.68244 18.68250 18.68257 18.68263 18.65823 18.65795 18.65768
#> 17 13.41460 13.41459 13.41457 13.41456 13.41455 13.41956 13.41967 13.41977
#> 18 16.55350 16.55358 16.55366 16.55374 16.55381 16.71984 16.71955 16.71927
#> 19 25.60492 25.60500 25.60508 25.60514 25.60520 25.46331 25.46361 25.46389
#> 20 17.82012 17.81999 17.81987 17.81976 17.81966 17.70951 17.70938 17.70926
#> 21 23.53564 23.53570 23.53576 23.53581 23.53585 23.49459 23.49470 23.49480
#>      0.0227   0.0212   0.0197   0.0184   0.0172   0.0160   0.0149   0.0139
#> 1  21.56578 21.56580 21.56581 21.56581 21.56581 21.56579 21.56577 21.56574
#> 2  24.05464 24.05463 24.05462 24.05461 24.05459 24.05457 24.05454 24.05452
#> 3  17.79977 17.79957 17.79939 17.79921 17.79904 17.79888 17.79872 17.79858
#> 4  18.97781 18.97766 18.97753 18.97739 18.97726 18.97714 18.97702 18.97691
#> 5  11.58047 11.58063 11.58076 11.58087 11.58096 11.58102 11.58107 11.58110
#> 6  17.82274 17.82258 17.82244 17.82231 17.82221 17.82211 17.82204 17.82197
#> 7  15.32542 15.32548 15.32553 15.32560 15.32567 15.32574 15.32582 15.32590
#> 8  16.54472 16.54480 16.54488 16.54496 16.54504 16.54511 16.54519 16.54526
#> 9  17.46012 17.46025 17.46037 17.46050 17.46062 17.46074 17.46086 17.46097
#> 10 12.66623 12.66622 12.66621 12.66619 12.66616 12.66613 12.66609 12.66605
#> 11 12.58770 12.58782 12.58793 12.58805 12.58816 12.58827 12.58837 12.58848
#> 12 28.33546 28.33543 28.33540 28.33538 28.33536 28.33534 28.33533 28.33532
#> 13 31.94697 31.94673 31.94650 31.94629 31.94610 31.94592 31.94575 31.94560
#> 14 31.12967 31.12969 31.12972 31.12974 31.12977 31.12979 31.12981 31.12984
#> 15 23.93802 23.93836 23.93868 23.93898 23.93927 23.93953 23.93977 23.94000
#> 16 18.65744 18.65721 18.65700 18.65681 18.65663 18.65647 18.65632 18.65619
#> 17 13.41986 13.41996 13.42004 13.42013 13.42021 13.42028 13.42036 13.42042
#> 18 16.71900 16.71874 16.71850 16.71827 16.71805 16.71784 16.71764 16.71745
#> 19 25.46416 25.46441 25.46466 25.46489 25.46511 25.46532 25.46552 25.46571
#> 20 17.70915 17.70905 17.70896 17.70888 17.70880 17.70873 17.70867 17.70861
#> 21 23.49489 23.49498 23.49506 23.49514 23.49521 23.49527 23.49533 23.49539
#>      0.0130   0.0121   0.0113   0.0105   0.0098   0.0092   0.0085   0.0080
#> 1  21.56571 21.56567 21.56563 21.56559 21.56554 21.56548 21.56543 21.56537
#> 2  24.05449 24.05445 24.05442 24.05439 24.05435 24.05431 24.05427 24.05424
#> 3  17.79844 17.79831 17.79818 17.79807 17.79796 17.79785 17.79776 17.79766
#> 4  18.97680 18.97670 18.97660 18.97650 18.97641 18.97633 18.97625 18.97617
#> 5  11.58111 11.58111 11.58110 11.58107 11.58103 11.58099 11.58093 11.58087
#> 6  17.82191 17.82187 17.82184 17.82181 17.82179 17.82178 17.82178 17.82179
#> 7  15.32598 15.32607 15.32615 15.32624 15.32633 15.32642 15.32650 15.32659
#> 8  16.54533 16.54540 16.54547 16.54553 16.54560 16.54566 16.54572 16.54578
#> 9  17.46108 17.46119 17.46129 17.46139 17.46149 17.46158 17.46167 17.46176
#> 10 12.66600 12.66595 12.66590 12.66585 12.66579 12.66573 12.66567 12.66561
#> 11 12.58858 12.58868 12.58878 12.58887 12.58896 12.58905 12.58913 12.58922
#> 12 28.33532 28.33531 28.33531 28.33532 28.33532 28.33533 28.33533 28.33534
#> 13 31.94546 31.94533 31.94521 31.94510 31.94500 31.94492 31.94484 31.94477
#> 14 31.12986 31.12988 31.12990 31.12992 31.12995 31.12997 31.12999 31.13001
#> 15 23.94021 23.94040 23.94058 23.94075 23.94090 23.94104 23.94116 23.94128
#> 16 18.65607 18.65596 18.65586 18.65578 18.65570 18.65563 18.65557 18.65552
#> 17 13.42049 13.42055 13.42061 13.42066 13.42071 13.42076 13.42081 13.42085
#> 18 16.71727 16.71711 16.71695 16.71680 16.71666 16.71653 16.71641 16.71629
#> 19 25.46589 25.46606 25.46621 25.46636 25.46650 25.46663 25.46675 25.46687
#> 20 17.70856 17.70851 17.70847 17.70844 17.70841 17.70838 17.70836 17.70835
#> 21 23.49544 23.49548 23.49552 23.49556 23.49559 23.49562 23.49565 23.49567
#>      0.0074   0.0069   0.0065   0.0060   0.0056   0.0052
#> 1  21.56532 21.56526 21.56520 21.56514 21.56508 21.56502
#> 2  24.05420 24.05416 24.05412 24.05408 24.05404 24.05400
#> 3  17.79758 17.79750 17.79742 17.79735 17.79729 17.79723
#> 4  18.97610 18.97603 18.97596 18.97590 18.97585 18.97579
#> 5  11.58080 11.58072 11.58064 11.58056 11.58047 11.58038
#> 6  17.82180 17.82181 17.82183 17.82186 17.82188 17.82191
#> 7  15.32668 15.32676 15.32685 15.32693 15.32701 15.32709
#> 8  16.54583 16.54588 16.54594 16.54599 16.54603 16.54608
#> 9  17.46184 17.46192 17.46200 17.46207 17.46214 17.46221
#> 10 12.66555 12.66549 12.66543 12.66537 12.66531 12.66525
#> 11 12.58929 12.58937 12.58944 12.58951 12.58958 12.58964
#> 12 28.33535 28.33536 28.33538 28.33539 28.33540 28.33542
#> 13 31.94470 31.94465 31.94460 31.94456 31.94452 31.94449
#> 14 31.13003 31.13005 31.13007 31.13008 31.13010 31.13012
#> 15 23.94138 23.94148 23.94156 23.94164 23.94170 23.94176
#> 16 18.65548 18.65544 18.65541 18.65539 18.65537 18.65535
#> 17 13.42089 13.42092 13.42096 13.42099 13.42102 13.42105
#> 18 16.71618 16.71608 16.71599 16.71590 16.71582 16.71575
#> 19 25.46697 25.46707 25.46716 25.46725 25.46732 25.46740
#> 20 17.70833 17.70832 17.70831 17.70831 17.70831 17.70831
#> 21 23.49569 23.49571 23.49573 23.49574 23.49575 23.49576
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 22.8 18.7 18.1 14.3 19.2 16.4 17.3 15.2 10.4 14.7 32.4 30.4 33.9 21.5
#> [16] 15.2 13.3 19.2 26.0 15.8 21.4
#> 
#> $X
#>               am       carb        cyl       disp       drat       gear
#>  [1,]  1.2747549  1.2126781 -0.2102353 -0.6614601  0.5037940  0.7179582
#>  [2,]  1.2747549 -1.3339459 -1.3139708 -1.0737636  0.4169329  0.7179582
#>  [3,] -0.7844645 -0.4850713  0.8935002  0.9243225 -0.7991214 -0.7897540
#>  [4,] -0.7844645 -1.3339459 -0.2102353 -0.1460808 -1.4766374 -0.7897540
#>  [5,] -0.7844645  1.2126781  0.8935002  0.9243225 -0.6948882 -0.7897540
#>  [6,] -0.7844645  1.2126781 -0.2102353 -0.6012004  0.5385384  0.7179582
#>  [7,] -0.7844645  0.3638034  0.8935002  0.2567080 -0.9380991 -0.7897540
#>  [8,] -0.7844645  0.3638034  0.8935002  0.2567080 -0.9380991 -0.7897540
#>  [9,] -0.7844645  0.3638034  0.8935002  0.2567080 -0.9380991 -0.7897540
#> [10,] -0.7844645  1.2126781  0.8935002  1.8123608 -1.1813100 -0.7897540
#> [11,] -0.7844645  1.2126781  0.8935002  1.5586355 -0.6601438 -0.7897540
#> [12,]  1.2747549 -1.3339459 -1.3139708 -1.3060808  0.8164936  0.7179582
#> [13,]  1.2747549 -0.4850713 -1.3139708 -1.3298675  2.2931311  0.7179582
#> [14,]  1.2747549 -1.3339459 -1.3139708 -1.3663405  1.0597045  0.7179582
#> [15,] -0.7844645 -1.3339459 -1.3139708 -0.9778238  0.1563498 -0.7897540
#> [16,] -0.7844645 -0.4850713  0.8935002  0.4803034 -0.7991214 -0.7897540
#> [17,] -0.7844645  1.2126781  0.8935002  0.8450334  0.2084665 -0.7897540
#> [18,] -0.7844645 -0.4850713  0.8935002  1.2414790 -0.9207269 -0.7897540
#> [19,]  1.2747549 -0.4850713 -1.3139708 -0.9762380  1.4245208  2.2256703
#> [20,]  1.2747549  1.2126781  0.8935002  0.8529623  1.0597045  2.2256703
#> [21,]  1.2747549 -0.4850713 -1.3139708 -0.9706878  0.8686103  0.7179582
#>                hp       qsec         vs         wt
#>  [1,] -0.62539091 -0.5079070 -0.7844645 -0.4079468
#>  [2,] -0.89438434  0.5897639  1.2747549 -0.9856649
#>  [3,]  0.40311342 -0.5079070 -0.7844645  0.1801806
#>  [4,] -0.70450662  1.7012419  1.2747549  0.2009993
#>  [5,]  1.51073345 -1.3225307 -0.7844645  0.3155020
#>  [6,] -0.41969004  0.3757526  1.2747549  0.1801806
#>  [7,]  0.48222913 -0.2455706 -0.7844645  0.8359687
#>  [8,]  0.48222913 -0.1074988 -0.7844645  0.4820513
#>  [9,]  0.48222913  0.1686448 -0.7844645  0.5340980
#> [10,]  0.87780772  0.1548377 -0.7844645  2.0642702
#> [11,]  1.27338630 -0.2317634 -0.7844645  2.1631589
#> [12,] -1.32160922  1.1834726  1.2747549 -1.1105769
#> [13,] -1.54313322  0.5276315  1.2747549 -1.7195229
#> [14,] -1.33743236  1.4803270  1.2747549 -1.4905176
#> [15,] -0.83109177  1.5562665  1.2747549 -0.8347295
#> [16,]  0.00753483 -0.3146065 -0.7844645  0.1749760
#> [17,]  1.51073345 -1.6193850 -0.7844645  0.5965540
#> [18,]  0.40311342 -0.4871962 -0.7844645  0.6017587
#> [19,] -0.92603063 -0.7288219 -0.7844645 -1.1730329
#> [20,]  1.81137318 -2.2476118 -0.7844645 -0.1008714
#> [21,] -0.64121405  0.5828603  1.2747549 -0.5068355
#> attr(,"center")
#>  [1]   0.3809524   2.5714286   6.3809524 243.4238095   3.6100000   3.5238095
#>  [7] 149.5238095  17.7557143   0.3809524   3.2669048
#> attr(,"scale")
#>  [1]   0.4856209   1.1780302   1.8120283 126.1206901   0.5756322   0.6632566
#>  [7]  63.1985677   1.4485216   0.4856209   0.9606762
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 39
#> 
#> $lambda.min
#> [1] 0.3700401
#> 
#> $null.dev
#> [1] 37.71365
#> 
#> $Bias
#> [1] 5.644071
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 3.731588