Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 42.637815 39.094096 32.907612 27.792067 23.590100 20.166052 17.403054
#>   [8] 15.200512 13.471905 12.142898 11.149685 10.437563  9.959695  9.676032
#>  [15]  9.552385  9.557798  9.526739  9.354930  9.045566  8.551614  8.132566
#>  [22]  7.623105  7.178794  6.807169  6.497765  6.241545  6.030716  5.858554
#>  [29]  4.684566  4.571107  4.483914  4.418943  4.355117  4.296635  4.255287
#>  [36]  4.219016  4.187633  4.187559  4.187498  4.299430  5.184942  5.316770
#>  [43]  5.315474  5.228952  5.254126  4.921092  5.691899  5.692142  5.692354
#>  [50]  4.851218  4.920646  4.920922  5.082934  5.730383  5.539148  5.418311
#>  [57]  5.972224  6.296732  6.008814  5.802877  5.879350  6.267741  6.267746
#>  [64]  6.167540  6.167488  6.159504  8.009578  8.159835  8.160265  7.697652
#>  [71]  7.719809  7.720050  7.720536  7.753165  7.753896  7.754603  7.732274
#>  [78]  7.732869  7.733449  7.734013  7.744780  7.678065  8.011836  8.012600
#>  [85]  8.013355  7.407851  7.408323  7.408800  7.409281  7.409768  7.410259
#>  [92]  7.410754  7.422089  7.422619  7.422251  7.423415  7.423854  7.277683
#>  [99]  7.278059  7.278428
#> 
#> $cvse
#>   [1] 12.305595 11.764463 10.152564  8.838828  7.765741  6.883605  6.150648
#>   [8]  5.533004  5.004407  4.545558  4.143264  3.789474  3.480319  3.215172
#>  [15]  2.995682  2.824617  2.689013  2.600791  2.602377  2.580868  2.567776
#>  [22]  2.415834  2.269584  2.142924  2.033015  1.937451  1.854196  1.781530
#>  [29]  1.504596  1.437898  1.380333  1.330885  1.288406  1.254230  1.225162
#>  [36]  1.200365  1.179489  1.179487  1.179485  1.178747  1.520750  1.471532
#>  [43]  1.425913  1.395745  1.387228  1.347739  1.492406  1.492451  1.492494
#>  [50]  1.320086  1.320583  1.320682  1.321249  1.442389  1.477195  1.462395
#>  [57]  1.707764  1.685827  1.709132  1.661254  1.626617  1.808669  1.808667
#>  [64]  1.777382  1.777371  1.777519  2.349141  2.341007  2.341450  2.234573
#>  [71]  2.170519  2.170802  2.171193  2.179801  2.180228  2.180648  2.181139
#>  [78]  2.181548  2.181950  2.182346  2.188515  2.167409  2.123816  2.124155
#>  [85]  2.124490  1.956107  1.956442  1.956773  1.957102  1.957427  1.957750
#>  [92]  1.958070  1.957329  1.957644  1.957936  1.958218  1.958514  1.941470
#>  [99]  1.941792  1.942110
#> 
#> $fold
#>  [1]  9  1 10  2  8  2  3  3  1  4  6  7  1  6  7 10  4  5  5  8  9
#> 
#> $lambda
#>   [1] 5.799684181 5.408804878 5.044269532 4.704302648 4.387248394 4.091562536
#>   [7] 3.815804915 3.558632435 3.318792519 3.095117011 2.886516484 2.691974933
#>  [13] 2.510544833 2.341342513 2.183543864 2.036380316 1.899135098 1.771139749
#>  [19] 1.651770857 1.540447030 1.436626056 1.339802268 1.249504079 1.165291686
#>  [25] 1.086754927 1.013511282 0.945204013 0.881500426 0.822090248 0.766684117
#>  [31] 0.715012173 0.666822746 0.621881125 0.579968418 0.540880488 0.504426953
#>  [37] 0.470430265 0.438724839 0.409156254 0.381580492 0.355863244 0.331879252
#>  [43] 0.309511701 0.288651648 0.269197493 0.251054483 0.234134251 0.218354386
#>  [49] 0.203638032 0.189913510 0.177113975 0.165177086 0.154044703 0.143662606
#>  [55] 0.133980227 0.124950408 0.116529169 0.108675493 0.101351128 0.094520401
#>  [61] 0.088150043 0.082209025 0.076668412 0.071501217 0.066682275 0.062188112
#>  [67] 0.057996842 0.054088049 0.050442695 0.047043026 0.043872484 0.040915625
#>  [73] 0.038158049 0.035586324 0.033187925 0.030951170 0.028865165 0.026919749
#>  [79] 0.025105448 0.023413425 0.021835439 0.020363803 0.018991351 0.017711397
#>  [85] 0.016517709 0.015404470 0.014366261 0.013398023 0.012495041 0.011652917
#>  [91] 0.010867549 0.010135113 0.009452040 0.008815004 0.008220902 0.007666841
#>  [97] 0.007150122 0.006668227 0.006218811 0.005799684
#> 
#> $fit
#> $beta
#>               5.7997     5.4088    5.0443    4.7043    4.3872    4.0916
#> (Intercept) 19.93333 21.7777622 23.497883 25.102073 26.598146 27.993389
#> am           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000 -0.5791245 -1.119218 -1.622911 -2.092657 -2.530743
#>                3.8158    3.5586    3.3188    3.0951    2.8865    2.6920
#> (Intercept) 29.294597 30.508108 31.639832 32.695282 33.679599 34.597575
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.939304 -3.320329 -3.675675 -4.007071 -4.316132 -4.604364
#>               2.5105    2.3413    2.1835    2.0364    1.8991      1.7711
#> (Intercept) 35.45368 36.252093 36.996692 37.691107 38.177848 37.87474566
#> am           0.00000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> carb         0.00000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> cyl          0.00000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> disp         0.00000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> drat         0.00000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> gear         0.00000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> hp           0.00000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> qsec         0.00000  0.000000  0.000000  0.000000  0.000000  0.01699922
#> vs           0.00000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> wt          -4.87317 -5.123859 -5.357653 -5.575689 -5.728519 -5.72582032
#>                 1.6518     1.5404     1.4366     1.3398     1.2495     1.1653
#> (Intercept) 35.6177574 33.5129664 31.5500312 29.7193912 28.0121303 26.4199331
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.1435802  0.2616254  0.3717147  0.4743844  0.5701344  0.6594313
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.7057244 -5.6869836 -5.6695059 -5.6532062 -5.6380049 -5.6238282
#>                 1.0868     1.0135     0.9452     0.8815    0.8221    0.7667
#> (Intercept) 24.9350446 23.5502326 22.2587523 21.0543135 19.931050 18.883490
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> qsec         0.7427098  0.8203756  0.8928071  0.9603568  1.023354  1.082105
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> wt          -5.6106070 -5.5982768 -5.5867776 -5.5760535 -5.566052 -5.556725
#>                0.7150    0.6668    0.6219    0.5800    0.5409    0.5044
#> (Intercept) 17.906533 16.995419 16.145711 15.793043 15.792989 15.792988
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         1.136897  1.187996  1.235651  1.255430  1.255433  1.255433
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -5.548026 -5.539914 -5.532348 -5.529208 -5.529207 -5.529207
#>                0.4704    0.4387    0.4092    0.3816    0.3559    0.3319
#> (Intercept) 15.792988 15.792988 15.792988 15.792988 15.792988 15.792988
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         1.255433  1.255433  1.255433  1.255433  1.255433  1.255433
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -5.529207 -5.529207 -5.529207 -5.529207 -5.529207 -5.529207
#>                  0.3095     0.2887     0.2692     0.2511     0.2341     0.2184
#> (Intercept)  8.50310795 10.2620687 11.6830534 11.6813130 11.6799562 11.6789026
#> am           2.03225917  2.4634411  2.8107993  2.8117030  2.8124022  2.8129427
#> carb        -0.04521142 -0.2890444 -0.4857518 -0.4858906 -0.4859973 -0.4860794
#> cyl          0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         1.48138657  1.3595637  1.2612193  1.2612435  1.2612627  1.2612778
#> vs           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -4.72797997 -4.4377908 -4.2038745 -4.2034558 -4.2031322 -4.2028822
#>                 0.2036     0.1899     0.1771     0.1652     0.1540     0.1437
#> (Intercept) 11.6780862 11.6774545 11.6769661 11.6765887 11.6762971 11.6760719
#> am           2.8133603  2.8136830  2.8139322  2.8141247  2.8142733  2.8143881
#> carb        -0.4861427 -0.4861915 -0.4862292 -0.4862583 -0.4862808 -0.4862981
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         1.2612896  1.2612988  1.2613059  1.2613114  1.2613157  1.2613190
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -4.2026891 -4.2025399 -4.2024248 -4.2023358 -4.2022672 -4.2022141
#>                 0.1340      0.1250      0.1165      0.1087      0.1014
#> (Intercept) 11.6758980 16.72171818 21.22653557 21.22298346 21.21932297
#> am           2.8144767  1.32845845  1.44998897  1.44987833  1.44983863
#> carb        -0.4863115  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.0000000 -1.12688434 -1.24595023 -1.24564264 -1.24534416
#> disp         0.0000000  0.02001291  0.01943606  0.01943231  0.01942922
#> drat         0.0000000  0.00000000 -0.69029679 -0.68995172 -0.68964222
#> gear         0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         1.2613215  1.34594288  1.28404355  1.28409615  1.28416344
#> vs           0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -4.2021732 -5.73114584 -5.74965835 -5.74955078 -5.74948976
#>                 0.0945     0.0882      0.0822      0.0767      0.0715
#> (Intercept) 21.2156440 12.2481593 12.24583006 12.24356200 12.24135746
#> am           1.4498562  0.0000000  0.00000000  0.00000000  0.00000000
#> carb         0.0000000  0.7848814  0.78506803  0.78524340  0.78540871
#> cyl         -1.2450585 -1.7106451 -1.71081827 -1.71097154 -1.71110819
#> disp         0.0194267  0.0343038  0.03430845  0.03431285  0.03431701
#> drat        -0.6893667 -0.5298075 -0.52991569 -0.53001370 -0.53010304
#> gear         0.0000000  0.0000000  0.00000000  0.00000000  0.00000000
#> hp           0.0000000  0.0000000  0.00000000  0.00000000  0.00000000
#> qsec         1.2842411  2.0730320  2.07325165  2.07346009  2.07365829
#> vs           0.0000000 -1.5773028 -1.57763796 -1.57794627 -1.57823145
#> wt          -5.7494664 -7.9883737 -7.98886411 -7.98933697 -7.98979279
#>                  0.0667      0.0622      0.0580      0.0541      0.0504
#> (Intercept) 12.23921777 12.23714350  9.31001055  9.31338843  9.31664104
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.78556500  0.78571314  0.56475152  0.56457271  0.56440120
#> cyl         -1.71123086 -1.71134171 -1.86878510 -1.86866903 -1.86855699
#> disp         0.03432097  0.03432473  0.02981038  0.02980565  0.02980111
#> drat        -0.53018498 -0.53026054 -0.44098365 -0.44090903 -0.44083748
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.01175831  0.01175736  0.01175643
#> qsec         2.07384713  2.07402731  2.23683573  2.23657311  2.23632031
#> vs          -1.57849658 -1.57874420 -2.01188674 -2.01154974 -2.01122460
#> wt          -7.99023210 -7.99065543 -7.70545929 -7.70486562 -7.70429563
#>                  0.0470      0.0439      0.0409      0.0382      0.0356
#> (Intercept)  9.31977252  9.32278691  9.32568812  9.32847998  9.05197838
#> am           0.00000000  0.00000000  0.00000000  0.00000000 -0.66727346
#> carb         0.56423672  0.56407900  0.56392778  0.56378280  0.69536820
#> cyl         -1.86844885 -1.86834451 -1.86824385 -1.86814675 -2.05166073
#> disp         0.02979675  0.02979257  0.02978856  0.02978471  0.03221222
#> drat        -0.44076891 -0.44070318 -0.44064020 -0.44057984 -0.34983954
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.01175552  0.01175462  0.01175374  0.01175287  0.01296590
#> qsec         2.23607701  2.23584289  2.23561762  2.23540092  2.33948288
#> vs          -2.01091095 -2.01060844 -2.01031674 -2.01003551 -2.40269539
#> wt          -7.70374844 -7.70322320 -7.70271910 -7.70223534 -8.14587560
#>                  0.0332      0.0310      0.0289      0.0269      0.0251
#> (Intercept)  9.05003608  9.04813228  9.04626623  9.04443717  9.04264438
#> am          -0.66773534 -0.66818807 -0.66863183 -0.66906680 -0.66949315
#> carb         0.69554780  0.69572382  0.69589635  0.69606544  0.69623118
#> cyl         -2.05183998 -2.05201567 -2.05218788 -2.05235669 -2.05252215
#> disp         0.03221619  0.03222009  0.03222392  0.03222766  0.03223134
#> drat        -0.34980649 -0.34977410 -0.34974233 -0.34971120 -0.34968067
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.01296716  0.01296841  0.01296962  0.01297082  0.01297199
#> qsec         2.33968677  2.33988661  2.34008249  2.34027448  2.34046266
#> vs          -2.40313866 -2.40357316 -2.40399905 -2.40441651 -2.40482570
#> wt          -8.14647493 -8.14706235 -8.14763810 -8.14820241 -8.14875552
#>                  0.0234      0.0218      0.0204      0.0190      0.0177
#> (Intercept)  9.04088715  9.03916478  9.03747656  9.03582184  9.03419995
#> am          -0.66991105 -0.67032066 -0.67072215 -0.67111568 -0.67150141
#> carb         0.69639362  0.69655283  0.69670887  0.69686181  0.69701172
#> cyl         -2.05268433 -2.05284329 -2.05299911 -2.05315183 -2.05330153
#> disp         0.03223493  0.03223846  0.03224192  0.03224531  0.03224863
#> drat        -0.34965074 -0.34962140 -0.34959263 -0.34956444 -0.34953679
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.01297314  0.01297426  0.01297537  0.01297645  0.01297751
#> qsec         2.34064711  2.34082791  2.34100511  2.34117880  2.34134904
#> vs          -2.40522679 -2.40561992 -2.40600527 -2.40638298 -2.40675320
#> wt          -8.14929763 -8.14982898 -8.15034976 -8.15086020 -8.15136050
#>                  0.0165      0.0154      0.0144      0.0134      0.0125
#> (Intercept)  9.03261024  9.03105207  9.02952483  9.02802789  9.02656067
#> am          -0.67187949 -0.67225007 -0.67261330 -0.67296932 -0.67331828
#> carb         0.69715864  0.69730264  0.69744378  0.69758211  0.69771770
#> cyl         -2.05344826 -2.05359208 -2.05373305 -2.05387122 -2.05400665
#> disp         0.03225188  0.03225507  0.03225820  0.03226127  0.03226427
#> drat        -0.34950969 -0.34948313 -0.34945708 -0.34943155 -0.34940652
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.01297855  0.01297957  0.01298057  0.01298155  0.01298251
#> qsec         2.34151590  2.34167945  2.34183975  2.34199687  2.34215087
#> vs          -2.40711609 -2.40747177 -2.40782041 -2.40816213 -2.40849708
#> wt          -8.15185086 -8.15233148 -8.15280254 -8.15326424 -8.15371677
#>                  0.0117      0.0109      0.0101      0.0095      0.0088
#> (Intercept)  9.02512257  9.02371302  9.02233145  9.02097731  9.01965006
#> am          -0.67366031 -0.67399556 -0.67432415 -0.67464622 -0.67496190
#> carb         0.69785058  0.69798083  0.69810849  0.69823361  0.69835624
#> cyl         -2.05413940 -2.05426951 -2.05439703 -2.05452203 -2.05464455
#> disp         0.03226721  0.03227010  0.03227293  0.03227570  0.03227842
#> drat        -0.34938199 -0.34935794 -0.34933437 -0.34931126 -0.34928860
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.01298345  0.01298437  0.01298527  0.01298616  0.01298703
#> qsec         2.34230182  2.34244976  2.34259477  2.34273690  2.34287621
#> vs          -2.40882538 -2.40914717 -2.40946257 -2.40977172 -2.41007473
#> wt          -8.15416031 -8.15459503 -8.15502111 -8.15543872 -8.15584804
#>                  0.0082      0.0077      0.0072      0.0067      0.0062
#> (Intercept)  9.01834916  9.01707409  9.01582434  9.01459941  9.01339881
#> am          -0.67527132 -0.67557459 -0.67587184 -0.67616319 -0.67644875
#> carb         0.69847643  0.69859423  0.69870969  0.69882286  0.69893378
#> cyl         -2.05476464 -2.05488234 -2.05499770 -2.05511078 -2.05522161
#> disp         0.03228108  0.03228369  0.03228625  0.03228875  0.03229121
#> drat        -0.34926640 -0.34924463 -0.34922329 -0.34920238 -0.34918187
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.01298788  0.01298872  0.01298954  0.01299034  0.01299112
#> qsec         2.34301275  2.34314658  2.34327775  2.34340631  2.34353233
#> vs          -2.41037173 -2.41066284 -2.41094816 -2.41122783 -2.41150194
#> wt          -8.15624922 -8.15664243 -8.15702782 -8.15740555 -8.15777577
#>                  0.0058
#> (Intercept)  9.01222205
#> am          -0.67672865
#> carb         0.69904249
#> cyl         -2.05533024
#> disp         0.03229362
#> drat        -0.34916178
#> gear         0.00000000
#> hp           0.01299189
#> qsec         2.34365583
#> vs          -2.41177061
#> wt          -8.15813864
#> 
#> $iter
#>   [1]   0   7   3   3   2   2   2   2   2   2   2   2   2   2   2   2   2   3
#>  [19]   3   3   3   3   3   3   3   3   3   3   3   3   3   3   3   2   1   1
#>  [37]   1   1   1   1   1   1 100  62  36   1   1   1   1   1   1   1   1   1
#>  [55]   1 101  62   1   1   1 236   1   1   1   1   1 131   1   1   1   1   1
#>  [73]   1 151   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.799684181 5.408804878 5.044269532 4.704302648 4.387248394 4.091562536
#>   [7] 3.815804915 3.558632435 3.318792519 3.095117011 2.886516484 2.691974933
#>  [13] 2.510544833 2.341342513 2.183543864 2.036380316 1.899135098 1.771139749
#>  [19] 1.651770857 1.540447030 1.436626056 1.339802268 1.249504079 1.165291686
#>  [25] 1.086754927 1.013511282 0.945204013 0.881500426 0.822090248 0.766684117
#>  [31] 0.715012173 0.666822746 0.621881125 0.579968418 0.540880488 0.504426953
#>  [37] 0.470430265 0.438724839 0.409156254 0.381580492 0.355863244 0.331879252
#>  [43] 0.309511701 0.288651648 0.269197493 0.251054483 0.234134251 0.218354386
#>  [49] 0.203638032 0.189913510 0.177113975 0.165177086 0.154044703 0.143662606
#>  [55] 0.133980227 0.124950408 0.116529169 0.108675493 0.101351128 0.094520401
#>  [61] 0.088150043 0.082209025 0.076668412 0.071501217 0.066682275 0.062188112
#>  [67] 0.057996842 0.054088049 0.050442695 0.047043026 0.043872484 0.040915625
#>  [73] 0.038158049 0.035586324 0.033187925 0.030951170 0.028865165 0.026919749
#>  [79] 0.025105448 0.023413425 0.021835439 0.020363803 0.018991351 0.017711397
#>  [85] 0.016517709 0.015404470 0.014366261 0.013398023 0.012495041 0.011652917
#>  [91] 0.010867549 0.010135113 0.009452040 0.008815004 0.008220902 0.007666841
#>  [97] 0.007150122 0.006668227 0.006218811 0.005799684
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 42
#> 
#> $loss
#>   [1] 835.80667 700.20631 586.75646 492.26947 413.99329 349.55336 296.90212
#>   [8] 254.27523 220.15362 193.23043 172.38245 156.64527 145.19177 137.31344
#>  [15] 132.40419 129.94636 129.44360 127.61675 114.79669 104.08551  95.17850
#>  [22]  87.81310  81.76281  76.83234  72.85349  69.68145  67.19169  65.27719
#>  [29]  63.84613  62.81974  62.13062  61.72111  61.54200  61.52514  61.52514
#>  [36]  61.52514  61.52514  61.52514  61.52514  61.52514  61.52514  61.52514
#>  [43]  54.04030  50.49640  49.61015  49.61014  49.61014  49.61013  49.61013
#>  [50]  49.61013  49.61013  49.61013  49.61013  49.61013  49.61013  37.52590
#>  [57]  36.58173  36.58172  36.58171  36.58170  33.03910  33.03908  33.03906
#>  [64]  33.03905  33.03904  33.03902  32.02540  32.02538  32.02536  32.02534
#>  [71]  32.02532  32.02531  32.02529  31.76889  31.76887  31.76885  31.76883
#>  [78]  31.76881  31.76879  31.76877  31.76876  31.76874  31.76873  31.76871
#>  [85]  31.76870  31.76869  31.76867  31.76866  31.76865  31.76864  31.76863
#>  [92]  31.76862  31.76861  31.76860  31.76859  31.76858  31.76857  31.76856
#>  [99]  31.76855  31.76855
#> 
#> $linear.predictors
#>      5.7997   5.4088   5.0443   4.7043   4.3872   4.0916   3.8158   3.5586
#> 1  19.93333 20.26046 20.56553 20.85005 21.11539 21.36284 21.59362 21.80885
#> 2  19.93333 20.11278 20.28013 20.43620 20.58176 20.71750 20.84410 20.96216
#> 3  19.93333 19.91588 19.89960 19.88441 19.87025 19.85705 19.84473 19.83325
#> 4  19.93333 19.78557 19.64777 19.51926 19.39941 19.28763 19.18339 19.08618
#> 5  19.93333 19.78557 19.64777 19.51926 19.39941 19.28763 19.18339 19.08618
#> 6  19.93333 19.78557 19.64777 19.51926 19.39941 19.28763 19.18339 19.08618
#> 7  19.93333 19.61763 19.32320 19.04862 18.79254 18.55372 18.33099 18.12328
#> 8  19.93333 19.58867 19.26724 18.96747 18.68790 18.42718 18.18403 17.95726
#> 9  19.93333 18.73736 17.62199 16.58179 15.61170 14.70699 13.86325 13.07638
#> 10 19.93333 18.63659 17.42724 16.29940 15.24758 14.26664 13.35181 12.49864
#> 11 19.93333 20.84248 21.69035 22.48107 23.21851 23.90624 24.54762 25.14578
#> 12 19.93333 20.71507 21.44412 22.12403 22.75812 23.34948 23.90097 24.41530
#> 13 19.93333 19.78847 19.65337 19.52737 19.40987 19.30029 19.19809 19.10278
#> 14 19.93333 19.55392 19.20009 18.87009 18.56234 18.27534 18.00767 17.75804
#> 15 19.93333 19.55103 19.19449 18.86198 18.55188 18.26268 17.99297 17.74144
#> 16 19.93333 20.65716 21.33220 21.96174 22.54886 23.09640 23.60704 24.08327
#> 17 19.93333 20.53844 21.10276 21.62904 22.11986 22.57760 23.00449 23.40260
#> 18 19.93333 20.90155 21.80451 22.64661 23.43196 24.16437 24.84743 25.48445
#> 19 19.93333 19.94194 19.94996 19.95745 19.96442 19.97093 19.97700 19.98266
#> 20 19.93333 20.17359 20.39765 20.60661 20.80149 20.98323 21.15272 21.31080
#> 21 19.93333 19.71029 19.50227 19.30828 19.12736 18.95864 18.80128 18.65453
#>      3.3188   3.0951   2.8865    2.6920    2.5105    2.3413    2.1835    2.0364
#> 1  22.00956 22.19676 22.37133 22.534141 22.685978 22.827582 22.959642 23.082801
#> 2  21.07227 21.17495 21.27072 21.360028 21.443320 21.520998 21.593440 21.661000
#> 3  19.82254 19.81255 19.80323 19.794545 19.786442 19.778886 19.771838 19.765266
#> 4  18.99551 18.91096 18.83210 18.758563 18.689979 18.626017 18.566367 18.510736
#> 5  18.99551 18.91096 18.83210 18.758563 18.689979 18.626017 18.566367 18.510736
#> 6  18.99551 18.91096 18.83210 18.758563 18.689979 18.626017 18.566367 18.510736
#> 7  17.92957 17.74891 17.58042 17.423297 17.276760 17.140098 17.012647 16.893786
#> 8  17.74578 17.54855 17.36462 17.193079 17.033101 16.883905 16.744765 16.615002
#> 9  12.34254 11.65816 11.01990 10.424664  9.869541  9.351832  8.869015  8.418738
#> 10 11.70297 10.96093 10.26890  9.623504  9.021610  8.460281  7.936784  7.448568
#> 11 25.70362 26.22386 26.70904 27.161527 27.583514 27.977060 28.344083 28.686369
#> 12 24.89497 25.34231 25.75950 26.148567 26.511417 26.849811 27.165399 27.459717
#> 13 19.01389 18.93099 18.85368 18.781585 18.714345 18.651637 18.593155 18.538614
#> 14 17.52524 17.30813 17.10565 16.916817 16.740711 16.576474 16.423305 16.280460
#> 15 17.50686 17.28809 17.08407 16.893795 16.716345 16.550854 16.396517 16.252582
#> 16 24.52740 24.94160 25.32788 25.688131 26.024100 26.337425 26.629634 26.902148
#> 17 23.77389 24.12015 24.44308 24.744236 25.025100 25.287034 25.531315 25.759132
#> 18 26.07854 26.63258 27.14929 27.631172 28.080577 28.499694 28.890563 29.255089
#> 19 19.98794 19.99287 19.99746 20.001741 20.005735 20.009459 20.012933 20.016172
#> 20 21.45821 21.59570 21.72391 21.843487 21.955003 22.059003 22.155994 22.246448
#> 21 18.51767 18.39004 18.27101 18.159995 18.056467 17.959916 17.869872 17.785896
#>       1.8991    1.7711    1.6518    1.5404    1.4366   1.3398    1.2495
#> 1  23.169128 23.152904 23.032090 22.919423 22.814350 22.71636 22.624970
#> 2  21.708356 21.702339 21.657535 21.615752 21.576786 21.54045 21.506554
#> 3  19.760659 19.796698 20.065053 20.315311 20.548703 20.76637 20.969358
#> 4  18.471743 18.467250 18.433801 18.402607 18.373515 18.34638 18.321081
#> 5  18.471743 18.489009 18.617583 18.737487 18.849310 18.95360 19.050853
#> 6  18.471743 18.499209 18.703732 18.894462 19.072339 19.23823 19.392934
#> 7  16.810472 16.816622 16.862417 16.905124 16.944953 16.98210 17.016738
#> 8  16.524046 16.537131 16.634563 16.725425 16.810163 16.88919 16.962891
#> 9   8.103123  8.119835  8.244277  8.360327  8.468555  8.56949  8.663622
#> 10  7.106361  7.120822  7.228508  7.328931  7.422587  7.50993  7.591387
#> 11 28.926290 28.942371 29.062118 29.173790 29.277935 29.37506 29.465642
#> 12 27.666016 27.706150 28.004999 28.283696 28.543610 28.78601 29.012066
#> 13 18.500385 18.500639 18.502532 18.504297 18.505943 18.50748 18.508909
#> 14 16.180335 16.149554 15.920347 15.706596 15.507252 15.32134 15.147963
#> 15 16.151692 16.148803 16.127290 16.107227 16.088516 16.07107 16.054793
#> 16 27.093164 27.116569 27.290847 27.453373 27.604945 27.74630 27.878132
#> 17 25.918817 25.905377 25.805297 25.711965 25.624924 25.54375 25.468045
#> 18 29.510599 29.498866 29.411502 29.330029 29.254047 29.18319 29.117101
#> 19 20.018443 19.970384 19.612524 19.278796 18.967560 18.67730 18.406604
#> 20 22.309850 22.277711 22.038394 21.815215 21.607078 21.41297 21.231940
#> 21 17.727035 17.681756 17.344592 17.030165 16.736930 16.46346 16.208415
#>       1.1653    1.0868    1.0135    0.9452    0.8815    0.8221    0.7667
#> 1  22.539742 22.460258 22.386130 22.316999 22.252527 22.192400 22.136326
#> 2  21.474947 21.445470 21.417980 21.392343 21.368433 21.346135 21.325340
#> 3  21.158669 21.335222 21.499875 21.653431 21.796639 21.930194 22.054748
#> 4  18.297484 18.275477 18.254954 18.235813 18.217963 18.201316 18.185790
#> 5  19.141556 19.226146 19.305035 19.378606 19.447220 19.511209 19.570885
#> 6  19.537215 19.671772 19.797260 19.914291 20.023434 20.125221 20.220148
#> 7  17.049044 17.079173 17.107271 17.133476 17.157914 17.180706 17.201961
#> 8  17.031625 17.095727 17.155508 17.211260 17.263254 17.311745 17.356967
#> 9   8.751409  8.833280  8.909633  8.980841  9.047249  9.109181  9.166940
#> 10  7.667354  7.738201  7.804273  7.865892  7.923358  7.976951  8.026933
#> 11 29.550118 29.628900 29.702372 29.770893 29.834796 29.894392 29.949971
#> 12 29.222891 29.419506 29.602870 29.773876 29.933356 30.082089 30.220797
#> 13 18.510244 18.511489 18.512650 18.513733 18.514743 18.515685 18.516563
#> 14 14.986269 14.835472 14.694838 14.563683 14.441367 14.327295 14.220911
#> 15 16.039617 16.025463 16.012263 15.999953 15.988472 15.977765 15.967780
#> 16 28.001076 28.115735 28.222667 28.322391 28.415394 28.502129 28.583019
#> 17 25.397443 25.331599 25.270193 25.212926 25.159518 25.109710 25.063259
#> 18 29.055469 28.997992 28.944388 28.894397 28.847775 28.804295 28.763746
#> 19 18.154151 17.918713 17.699142 17.494370 17.303398 17.125298 16.959200
#> 20 21.063114 20.905665 20.758828 20.621888 20.494176 20.375072 20.263996
#> 21 15.970563 15.748741 15.541869 15.348939 15.169012 15.001212 14.844721
#>       0.7150    0.6668    0.6219    0.5800    0.5409    0.5044    0.4704
#> 1  22.084030 22.035259 21.989776 21.970898 21.970895 21.970895 21.970895
#> 2  21.305946 21.287859 21.270992 21.263991 21.263990 21.263990 21.263990
#> 3  22.170908 22.279239 22.380269 22.422201 22.422207 22.422207 22.422207
#> 4  18.171311 18.157808 18.145215 18.139988 18.139987 18.139987 18.139987
#> 5  19.626539 19.678443 19.726848 19.746939 19.746942 19.746942 19.746942
#> 6  20.308678 20.391240 20.468239 20.500197 20.500202 20.500202 20.500202
#> 7  17.221784 17.240271 17.257512 17.264667 17.264668 17.264668 17.264668
#> 8  17.399141 17.438473 17.475155 17.490379 17.490381 17.490381 17.490381
#> 9   9.220805  9.271040  9.317890  9.337335  9.337338  9.337338  9.337338
#> 10  8.073545  8.117016  8.157557  8.174384  8.174386  8.174386  8.174386
#> 11 30.001804 30.050144 30.095226 30.113938 30.113940 30.113941 30.113941
#> 12 30.350157 30.470798 30.583308 30.630005 30.630013 30.630013 30.630013
#> 13 18.517383 18.518146 18.518859 18.519155 18.519155 18.519155 18.519155
#> 14 14.121696 14.029169 13.942878 13.907063 13.907057 13.907057 13.907057
#> 15 15.958468 15.949783 15.941684 15.938322 15.938321 15.938321 15.938321
#> 16 28.658457 28.728811 28.794423 28.821655 28.821659 28.821659 28.821659
#> 17 25.019938 24.979537 24.941859 24.926221 24.926218 24.926218 24.926218
#> 18 28.725930 28.690662 28.657771 28.644120 28.644118 28.644118 28.644118
#> 19 16.804298 16.659835 16.525108 16.469190 16.469182 16.469182 16.469182
#> 20 20.160405 20.063796 19.973698 19.936304 19.936298 19.936298 19.936298
#> 21 14.698777 14.562669 14.435734 14.383050 14.383042 14.383042 14.383042
#>       0.4387    0.4092    0.3816    0.3559    0.3319    0.3095    0.2887
#> 1  21.970895 21.970895 21.970895 21.970895 21.970895 22.350837 22.320738
#> 2  21.263990 21.263990 21.263990 21.263990 21.263990 21.974778 21.950457
#> 3  22.422207 22.422207 22.422207 22.422207 22.422207 22.055596 22.135444
#> 4  18.139987 18.139987 18.139987 18.139987 18.139987 17.361633 17.557753
#> 5  19.746942 19.746942 19.746942 19.746942 19.746942 19.167385 18.719905
#> 6  20.500202 20.500202 20.500202 20.500202 20.500202 20.056217 19.535644
#> 7  17.264668 17.264668 17.264668 17.264668 17.264668 16.804512 16.770296
#> 8  17.490381 17.490381 17.490381 17.490381 17.490381 17.160668 17.092232
#> 9   9.337338  9.337338  9.337338  9.337338  9.337338 10.135698 10.252444
#> 10  8.174386  8.174386  8.174386  8.174386  8.174386  9.076008  9.262738
#> 11 30.113941 30.113941 30.113941 30.113941 30.113941 30.244536 30.159508
#> 12 30.630013 30.630013 30.630013 30.630013 30.630013 31.293905 31.348436
#> 13 18.519155 18.519155 18.519155 18.519155 18.519155 17.800062 17.960620
#> 14 13.907057 13.907057 13.907057 13.907057 13.907057 12.994986 13.015650
#> 15 15.938321 15.938321 15.938321 15.938321 15.938321 15.491243 15.801234
#> 16 28.821659 28.821659 28.821659 28.821659 28.821659 29.339721 29.545093
#> 17 24.926218 24.926218 24.926218 24.926218 24.926218 25.066223 25.355262
#> 18 28.644118 28.644118 28.644118 28.644118 28.644118 28.326944 28.409669
#> 19 16.469182 16.469182 16.469182 16.469182 16.469182 16.846930 17.215208
#> 20 19.936298 19.936298 19.936298 19.936298 19.936298 20.129086 19.771799
#> 21 14.383042 14.383042 14.383042 14.383042 14.383042 14.923031 14.419870
#>       0.2692    0.2511   0.2341    0.2184    0.2036    0.1899    0.1771
#> 1  22.296363 22.296467 22.29655 22.296610 22.296658 22.296695 22.296724
#> 2  21.930658 21.930882 21.93106 21.931190 21.931294 21.931375 21.931437
#> 3  22.199947 22.199885 22.19984 22.199798 22.199769 22.199746 22.199728
#> 4  17.716173 17.716008 17.71588 17.715778 17.715700 17.715640 17.715593
#> 5  18.359030 18.358618 18.35830 18.358055 18.357865 18.357719 18.357606
#> 6  19.115762 19.115364 19.11506 19.114822 19.114639 19.114499 19.114390
#> 7  16.742805 16.742636 16.74251 16.742404 16.742326 16.742265 16.742218
#> 8  17.037099 17.036961 17.03685 17.036771 17.036707 17.036658 17.036619
#> 9  10.346427 10.346765 10.34703 10.347229 10.347386 10.347506 10.347600
#> 10  9.413158  9.413565  9.41388  9.414123  9.414311  9.414457  9.414569
#> 11 30.090872 30.090883 30.09089 30.090897 30.090903 30.090907 30.090910
#> 12 31.392254 31.392529 31.39274 31.392906 31.393033 31.393132 31.393208
#> 13 18.090334 18.090173 18.09005 18.089950 18.089875 18.089816 18.089771
#> 14 13.032557 13.032242 13.03200 13.031809 13.031663 13.031550 13.031462
#> 15 16.051441 16.051445 16.05145 16.051449 16.051450 16.051450 16.051450
#> 16 29.710648 29.710940 29.71117 29.711340 29.711475 29.711579 29.711659
#> 17 25.588419 25.588605 25.58875 25.588859 25.588944 25.589009 25.589060
#> 18 28.476492 28.476421 28.47636 28.476321 28.476288 28.476261 28.476241
#> 19 17.512243 17.512529 17.51275 17.512920 17.513051 17.513153 17.513231
#> 20 19.483508 19.483374 19.48327 19.483192 19.483131 19.483084 19.483048
#> 21 14.013808 14.013709 14.01363 14.013577 14.013534 14.013500 14.013475
#>       0.1652    0.1540    0.1437    0.1340    0.1250    0.1165    0.1087
#> 1  22.296746 22.296763 22.296777 22.296787 21.629553 21.689688 21.689764
#> 2  21.931485 21.931522 21.931551 21.931573 20.921839 20.942589 20.942723
#> 3  22.199715 22.199704 22.199696 22.199690 22.863238 23.115880 23.115637
#> 4  17.715557 17.715529 17.715508 17.715491 18.104096 18.157078 18.156989
#> 5  18.357519 18.357452 18.357400 18.357360 18.230188 18.021527 18.021878
#> 6  19.114306 19.114241 19.114191 19.114153 19.037754 18.791953 18.792335
#> 7  16.742182 16.742154 16.742133 16.742116 15.537624 15.653130 15.653391
#> 8  17.036590 17.036567 17.036550 17.036536 15.789443 15.879264 15.879552
#> 9  10.347672 10.347727 10.347770 10.347804 11.264272 11.311583 11.311243
#> 10  9.414655  9.414722  9.414774  9.414814  9.811547  9.824142  9.823882
#> 11 30.090912 30.090914 30.090915 30.090916 30.728678 30.255659 30.255792
#> 12 31.393266 31.393312 31.393346 31.393373 31.233168 31.163419 31.163420
#> 13 18.089736 18.089709 18.089688 18.089672 17.388893 17.456939 17.457074
#> 14 13.031395 13.031343 13.031303 13.031271 13.444540 13.195172 13.195279
#> 15 16.051451 16.051451 16.051451 16.051451 16.623876 16.692751 16.692533
#> 16 29.711721 29.711769 29.711806 29.711835 29.472212 29.554596 29.554478
#> 17 25.589099 25.589129 25.589152 25.589170 26.162786 26.112126 26.111880
#> 18 28.476226 28.476214 28.476204 28.476197 29.521078 29.939777 29.939341
#> 19 17.513291 17.513338 17.513374 17.513402 17.408071 17.010143 17.010185
#> 20 19.483020 19.482999 19.482982 19.482969 19.177583 19.496299 19.496301
#> 21 14.013455 14.013439 14.013428 14.013418 14.249562 14.336283 14.336326
#>       0.1014    0.0945    0.0882    0.0822    0.0767    0.0715    0.0667
#> 1  21.689834 21.689898 21.738741 21.738772 21.738799 21.738822 21.738842
#> 2  20.942846 20.942959 20.862603 20.862633 20.862656 20.862674 20.862687
#> 3  23.115426 23.115244 23.027562 23.027605 23.027649 23.027692 23.027733
#> 4  18.156932 18.156902 18.616236 18.616280 18.616331 18.616387 18.616447
#> 5  18.022144 18.022338 17.675463 17.675195 17.674941 17.674699 17.674469
#> 6  18.792642 18.792883 18.919282 18.919146 18.919017 18.918894 18.918778
#> 7  15.653626 15.653838 15.440851 15.440685 15.440533 15.440393 15.440265
#> 8  15.879817 15.880061 15.870646 15.870542 15.870450 15.870367 15.870292
#> 9  11.310946 11.310690 11.675736 11.676022 11.676282 11.676520 11.676739
#> 10  9.823644  9.823428  9.505342  9.505444  9.505528  9.505599  9.505659
#> 11 30.255922 30.256047 30.874215 30.874326 30.874433 30.874537 30.874637
#> 12 31.163451 31.163504 31.411041 31.411216 31.411381 31.411537 31.411685
#> 13 17.457209 17.457341 17.315613 17.315461 17.315327 17.315208 17.315101
#> 14 13.195349 13.195387 13.002741 13.002499 13.002276 13.002068 13.001873
#> 15 16.692358 16.692218 16.852374 16.852420 16.852468 16.852519 16.852570
#> 16 29.554379 29.554299 28.884345 28.884303 28.884261 28.884218 28.884176
#> 17 26.111626 26.111372 26.279556 26.279607 26.279640 26.279660 26.279668
#> 18 29.938937 29.938565 29.610787 29.610808 29.610825 29.610839 29.610850
#> 19 17.010261 17.010363 16.243190 16.243029 16.242889 16.242766 16.242658
#> 20 19.496275 19.496228 19.753926 19.754007 19.754075 19.754132 19.754181
#> 21 14.336377 14.336435 15.039749 15.040001 15.040241 15.040469 15.040687
#>       0.0622    0.0580    0.0541    0.0504    0.0470   0.0439    0.0409
#> 1  21.738859 21.329557 21.329579 21.329600 21.329621 21.32964 21.329662
#> 2  20.862697 20.617293 20.617319 20.617344 20.617368 20.61739 20.617415
#> 3  23.027774 22.987462 22.987403 22.987346 22.987291 22.98724 22.987189
#> 4  18.616509 18.453740 18.453627 18.453518 18.453415 18.45332 18.453220
#> 5  17.674251 17.485569 17.485884 17.486188 17.486480 17.48676 17.487031
#> 6  18.918667 18.827670 18.827828 18.827980 18.828126 18.82827 18.828402
#> 7  15.440147 15.665309 15.665425 15.665535 15.665642 15.66574 15.665841
#> 8  15.870225 16.174770 16.174810 16.174849 16.174885 16.17492 16.174952
#> 9  11.676943 11.572252 11.572029 11.571814 11.571608 11.57141 11.571219
#> 10  9.505709  9.602598  9.602573  9.602548  9.602524  9.60250  9.602477
#> 11 30.874734 30.628396 30.628273 30.628155 30.628042 30.62793 30.627831
#> 12 31.411825 31.784105 31.783886 31.783674 31.783470 31.78327 31.783083
#> 13 17.315005 17.155242 17.155341 17.155437 17.155529 17.15562 17.155704
#> 14 13.001691 13.168961 13.169175 13.169380 13.169577 13.16977 13.169949
#> 15 16.852622 16.623418 16.623343 16.623271 16.623203 16.62314 16.623075
#> 16 28.884134 29.085722 29.085775 29.085827 29.085875 29.08592 29.085966
#> 17 26.279668 26.532484 26.532529 26.532570 26.532610 26.53265 26.532682
#> 18 29.610858 29.597798 29.597804 29.597810 29.597815 29.59782 29.597826
#> 19 16.242562 16.333234 16.333303 16.333370 16.333433 16.33350 16.333554
#> 20 19.754223 19.596489 19.596482 19.596477 19.596471 19.59647 19.596461
#> 21 15.040896 15.377929 15.377613 15.377309 15.377016 15.37673 15.376462
#>       0.0382    0.0356    0.0332    0.0310    0.0289    0.0269    0.0251
#> 1  21.329682 21.237736 21.237666 21.237597 21.237529 21.237462 21.237397
#> 2  20.617437 20.470648 20.470539 20.470432 20.470327 20.470224 20.470124
#> 3  22.987141 22.984738 22.984761 22.984784 22.984806 22.984828 22.984849
#> 4  18.453129 18.589050 18.589200 18.589347 18.589491 18.589632 18.589770
#> 5  17.487291 17.533718 17.533597 17.533478 17.533362 17.533248 17.533137
#> 6  18.828532 18.937407 18.937409 18.937410 18.937412 18.937413 18.937414
#> 7  15.665934 15.659563 15.659505 15.659449 15.659394 15.659340 15.659287
#> 8  16.174983 16.188062 16.188056 16.188050 16.188045 16.188039 16.188034
#> 9  11.571036 11.555365 11.555461 11.555556 11.555649 11.555739 11.555828
#> 10  9.602454  9.482289  9.482216  9.482144  9.482074  9.482005  9.481937
#> 11 30.627732 30.725719 30.725852 30.725982 30.726110 30.726235 30.726358
#> 12 31.782900 31.735511 31.735589 31.735665 31.735739 31.735812 31.735884
#> 13 17.155787 17.156803 17.156759 17.156715 17.156672 17.156630 17.156589
#> 14 13.170123 13.337453 13.337462 13.337471 13.337479 13.337488 13.337497
#> 15 16.623015 16.673132 16.673203 16.673271 16.673339 16.673405 16.673469
#> 16 29.086009 28.897861 28.897703 28.897547 28.897395 28.897246 28.897100
#> 17 26.532714 26.711226 26.711326 26.711425 26.711522 26.711617 26.711710
#> 18 29.597832 29.588287 29.588267 29.588247 29.588228 29.588209 29.588190
#> 19 16.333611 16.106129 16.105937 16.105748 16.105563 16.105382 16.105204
#> 20 19.596456 19.618243 19.618259 19.618275 19.618291 19.618307 19.618322
#> 21 15.376201 15.411059 15.411234 15.411406 15.411574 15.411739 15.411900
#>       0.0234    0.0218    0.0204   0.0190    0.0177    0.0165   0.0154
#> 1  21.237333 21.237271 21.237209 21.23715 21.237090 21.237032 21.23698
#> 2  20.470025 20.469928 20.469833 20.46974 20.469649 20.469559 20.46947
#> 3  22.984870 22.984891 22.984911 22.98493 22.984950 22.984968 22.98499
#> 4  18.589906 18.590039 18.590169 18.59030 18.590421 18.590544 18.59066
#> 5  17.533027 17.532920 17.532815 17.53271 17.532611 17.532512 17.53242
#> 6  18.937416 18.937417 18.937418 18.93742 18.937421 18.937422 18.93742
#> 7  15.659235 15.659184 15.659134 15.65908 15.659037 15.658990 15.65894
#> 8  16.188029 16.188023 16.188018 16.18801 16.188008 16.188004 16.18800
#> 9  11.555915 11.556001 11.556085 11.55617 11.556247 11.556326 11.55640
#> 10  9.481871  9.481806  9.481742  9.48168  9.481619  9.481559  9.48150
#> 11 30.726478 30.726596 30.726711 30.72682 30.726935 30.727044 30.72715
#> 12 31.735954 31.736023 31.736090 31.73616 31.736221 31.736284 31.73635
#> 13 17.156549 17.156509 17.156470 17.15643 17.156395 17.156358 17.15632
#> 14 13.337505 13.337513 13.337521 13.33753 13.337537 13.337544 13.33755
#> 15 16.673533 16.673595 16.673655 16.67372 16.673774 16.673831 16.67389
#> 16 28.896957 28.896817 28.896679 28.89654 28.896412 28.896283 28.89616
#> 17 26.711801 26.711891 26.711979 26.71206 26.712149 26.712231 26.71231
#> 18 29.588172 29.588154 29.588136 29.58812 29.588102 29.588086 29.58807
#> 19 16.105030 16.104859 16.104692 16.10453 16.104367 16.104210 16.10406
#> 20 19.618337 19.618351 19.618366 19.61838 19.618393 19.618407 19.61842
#> 21 15.412059 15.412214 15.412366 15.41252 15.412662 15.412805 15.41295
#>       0.0144    0.0134    0.0125    0.0117    0.0109    0.0101   0.0095
#> 1  21.236920 21.236866 21.236812 21.236760 21.236709 21.236658 21.23661
#> 2  20.469386 20.469301 20.469219 20.469138 20.469059 20.468981 20.46890
#> 3  22.985005 22.985023 22.985040 22.985057 22.985074 22.985090 22.98511
#> 4  18.590782 18.590897 18.591010 18.591121 18.591230 18.591336 18.59144
#> 5  17.532321 17.532227 17.532136 17.532047 17.531959 17.531873 17.53179
#> 6  18.937424 18.937426 18.937427 18.937428 18.937429 18.937430 18.93743
#> 7  15.658899 15.658855 15.658811 15.658769 15.658727 15.658686 15.65865
#> 8  16.187995 16.187990 16.187986 16.187981 16.187977 16.187973 16.18797
#> 9  11.556479 11.556553 11.556626 11.556697 11.556767 11.556836 11.55690
#> 10  9.481442  9.481386  9.481331  9.481276  9.481223  9.481171  9.48112
#> 11 30.727255 30.727357 30.727457 30.727555 30.727652 30.727746 30.72784
#> 12 31.736407 31.736467 31.736525 31.736583 31.736639 31.736694 31.73675
#> 13 17.156288 17.156253 17.156219 17.156186 17.156154 17.156122 17.15609
#> 14 13.337559 13.337566 13.337573 13.337580 13.337587 13.337593 13.33760
#> 15 16.673942 16.673996 16.674049 16.674101 16.674151 16.674201 16.67425
#> 16 28.896032 28.895910 28.895790 28.895673 28.895558 28.895446 28.89534
#> 17 26.712392 26.712469 26.712546 26.712621 26.712694 26.712766 26.71284
#> 18 29.588054 29.588038 29.588023 29.588008 29.587994 29.587979 29.58797
#> 19 16.103904 16.103756 16.103610 16.103468 16.103328 16.103191 16.10306
#> 20 19.618433 19.618446 19.618458 19.618470 19.618482 19.618494 19.61851
#> 21 15.413083 15.413218 15.413351 15.413480 15.413607 15.413732 15.41385
#>      0.0088    0.0082    0.0077    0.0072   0.0067    0.0062   0.0058
#> 1  21.23656 21.236513 21.236467 21.236421 21.23638 21.236333 21.23629
#> 2  20.46883 20.468757 20.468685 20.468615 20.46855 20.468478 20.46841
#> 3  22.98512 22.985138 22.985153 22.985168 22.98518 22.985196 22.98521
#> 4  18.59154 18.591643 18.591742 18.591838 18.59193 18.592025 18.59212
#> 5  17.53171 17.531625 17.531546 17.531468 17.53139 17.531318 17.53124
#> 6  18.93743 18.937433 18.937434 18.937435 18.93744 18.937437 18.93744
#> 7  15.65861 15.658569 15.658531 15.658494 15.65846 15.658423 15.65839
#> 8  16.18797 16.187961 16.187958 16.187954 16.18795 16.187947 16.18794
#> 9  11.55697 11.557033 11.557096 11.557158 11.55722 11.557279 11.55734
#> 10  9.48107  9.481021  9.480973  9.480926  9.48088  9.480835  9.48079
#> 11 30.72793 30.728018 30.728105 30.728191 30.72827 30.728357 30.72844
#> 12 31.73680 31.736853 31.736904 31.736954 31.73700 31.737051 31.73710
#> 13 17.15606 17.156031 17.156001 17.155973 17.15594 17.155917 17.15589
#> 14 13.33761 13.337612 13.337619 13.337624 13.33763 13.337636 13.33764
#> 15 16.67430 16.674345 16.674390 16.674435 16.67448 16.674523 16.67457
#> 16 28.89523 28.895121 28.895017 28.894916 28.89482 28.894718 28.89462
#> 17 26.71291 26.712973 26.713039 26.713104 26.71317 26.713230 26.71329
#> 18 29.58795 29.587938 29.587925 29.587912 29.58790 29.587886 29.58787
#> 19 16.10293 16.102797 16.102670 16.102547 16.10243 16.102306 16.10219
#> 20 19.61852 19.618527 19.618538 19.618549 19.61856 19.618569 19.61858
#> 21 15.41397 15.414091 15.414206 15.414318 15.41443 15.414537 15.41464
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 21.4 18.7 19.2 17.8 17.3 15.2 10.4 10.4 30.4 33.9 15.2 13.3 19.2
#> [16] 27.3 26.0 30.4 15.8 19.7 15.0
#> 
#> $X
#>               am       carb        cyl       disp        drat       gear
#>  [1,]  1.0488088  0.4819908 -0.2936101 -0.6451061  0.45670980  0.2399041
#>  [2,]  1.0488088  0.4819908 -0.2936101 -0.6451061  0.45670980  0.2399041
#>  [3,] -0.9534626 -1.3042103 -0.2936101  0.1409379 -1.05861636 -1.0195922
#>  [4,] -0.9534626 -0.7088100  0.9395524  0.9590653 -0.92925924 -1.0195922
#>  [5,] -0.9534626  0.4819908 -0.2936101 -0.5841476  0.49366897  0.2399041
#>  [6,] -0.9534626  0.4819908 -0.2936101 -0.5841476  0.49366897  0.2399041
#>  [7,] -0.9534626 -0.1134096  0.9395524  0.2837092 -1.07709594 -1.0195922
#>  [8,] -0.9534626 -0.1134096  0.9395524  0.2837092 -1.07709594 -1.0195922
#>  [9,] -0.9534626  0.4819908  0.9395524  1.8574013 -1.33581016 -1.0195922
#> [10,] -0.9534626  0.4819908  0.9395524  1.7611510 -1.20645305 -1.0195922
#> [11,]  1.0488088 -0.7088100 -1.5267726 -1.3212643  2.36010729  0.2399041
#> [12,]  1.0488088 -1.3042103 -1.5267726 -1.3581603  1.04805659  0.2399041
#> [13,] -0.9534626 -0.7088100  0.9395524  0.5098973 -0.92925924 -1.0195922
#> [14,] -0.9534626  0.4819908  0.9395524  0.8788568  0.14255682 -1.0195922
#> [15,] -0.9534626 -0.7088100  0.9395524  1.2798996 -1.05861636 -1.0195922
#> [16,]  1.0488088 -1.3042103 -1.5267726 -1.2947955  0.78934237  0.2399041
#> [17,]  1.0488088 -0.7088100 -1.5267726 -0.9635341  1.43612792  1.4994004
#> [18,]  1.0488088 -0.7088100 -1.5267726 -1.1656597  0.21647516  1.4994004
#> [19,]  1.0488088  0.4819908  0.9395524  0.8868776  1.04805659  1.4994004
#> [20,]  1.0488088  1.6727915 -0.2936101 -0.7654189 -0.06071864  1.4994004
#> [21,]  1.0488088  2.8635923  0.9395524  0.4858348 -0.20855534  1.4994004
#>                hp        qsec         vs          wt
#>  [1,] -0.65185605 -0.59991412 -0.7071068 -0.55792606
#>  [2,] -0.65185605 -0.21142348 -0.7071068 -0.30605504
#>  [3,] -0.65185605  1.46741109  1.4142136  0.02977299
#>  [4,]  0.28279049 -0.21142348 -0.7071068  0.25201212
#>  [5,] -0.46492674  0.67655513  1.4142136  0.25201212
#>  [6,] -0.46492674  1.09279511  1.4142136  0.25201212
#>  [7,]  0.35468638  0.19094183 -0.7071068  0.53845368
#>  [8,]  0.35468638  0.46843515 -0.7071068  0.58784015
#>  [9,]  0.71416582  0.45456048 -0.7071068  2.03980250
#> [10,]  0.85795760  0.34356315 -0.7071068  2.21166744
#> [11,] -1.48584835  0.82917646  1.4142136 -1.55059420
#> [12,] -1.29891904  1.78652840  1.4142136 -1.33329371
#> [13,] -0.07668895 -0.01717816 -0.7071068  0.24707348
#> [14,]  1.28933292 -1.32833408 -0.7071068  0.64710392
#> [15,]  0.28279049 -0.19061148 -0.7071068  0.65204257
#> [16,] -1.28453987  1.09279511  1.4142136 -1.23452076
#> [17,] -0.92506043 -0.43341813 -0.7071068 -1.03203621
#> [18,] -0.60871852 -0.29467147  1.4142136 -1.65134261
#> [19,]  1.56253730 -1.95963137 -0.7071068 -0.01467484
#> [20,]  0.28279049 -1.26589808 -0.7071068 -0.40976663
#> [21,]  2.58345891 -1.89025804 -0.7071068  0.38041696
#> attr(,"center")
#>  [1]   0.4761905   3.1904762   6.4761905 240.4285714   3.6528571   3.8095238
#>  [7] 155.3333333  17.3247619   0.3333333   3.1848571
#> attr(,"scale")
#>  [1]   0.4994328   1.6795421   1.6218463 124.6749570   0.5411376   0.7939682
#>  [7]  69.5450066   1.4414762   0.4714045   1.0124229
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 39
#> 
#> $lambda.min
#> [1] 0.4091563
#> 
#> $null.dev
#> [1] 39.80032
#> 
#> $Bias
#> [1] 1.790742
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 14.32967