Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 47.965086 45.071138 40.087761 35.172291 31.131838 27.836606 25.174775
#>   [8] 23.050081 21.379727 20.092565 19.127521 18.432230 17.961845 17.678020
#>  [15] 17.548010 17.383232 16.413407 15.679766 15.223363 14.887013 14.595033
#>  [22] 14.341609 14.121685 13.959419 13.824812 13.747470 13.930305 13.044293
#>  [29] 11.572899 10.645478 10.692645 10.766491  9.627541 10.197700 10.254528
#>  [36] 10.315925 10.359684 10.282292 10.493633 10.675875 10.493865  9.893181
#>  [43]  9.874272 10.146396 10.097749  9.526483  8.720824  8.460509  8.541808
#>  [50]  8.291649  8.159200  7.219515  7.219595  7.545554  6.561176  6.560994
#>  [57]  6.422003  7.386205  7.343330  7.375292  7.375163  7.375026  7.893081
#>  [64]  7.987457  8.182808  8.182986  8.183158  7.811262  8.219508  8.186738
#>  [71]  8.186966  8.187189  8.187365  8.491750  8.492060  8.494284  9.303347
#>  [78]  9.551572  9.551846 10.449215 10.450088 10.450958 10.451824 10.490025
#>  [85] 10.490961 10.491885 10.492768 10.514149 10.515047 10.516004 10.516937
#>  [92] 10.523977 10.524939 10.565306 10.566308 10.567286 10.568241 10.569173
#>  [99] 10.570083 10.570972
#> 
#> $cvse
#>   [1] 13.527869 13.445007 12.406041 11.187773 10.092766  9.106491  8.218494
#>   [8]  7.421805  6.712430  6.088898  5.551758  5.102889  4.744508  4.477834
#>  [15]  4.301618  4.148534  3.857404  3.639461  3.524423  3.452528  3.394009
#>  [22]  3.345734  3.305316  3.279745  3.268164  3.253873  3.197792  3.181749
#>  [29]  2.875153  2.781389  2.771333  2.772654  2.799313  2.920447  2.919076
#>  [36]  2.919024  2.918729  2.911308  2.901124  2.809887  2.744939  2.429674
#>  [43]  2.426185  2.655696  2.661734  2.508344  2.281129  2.206844  2.216912
#>  [50]  2.063495  2.075766  2.114211  2.114219  2.144220  1.723997  1.723942
#>  [57]  1.679147  1.934272  1.923682  1.916103  1.916140  1.916172  2.348868
#>  [64]  2.355149  2.395966  2.396165  2.396360  2.402713  2.506426  2.503644
#>  [71]  2.503869  2.504088  2.504356  2.392181  2.392254  2.392602  2.435328
#>  [78]  2.430340  2.430399  2.682716  2.683025  2.683330  2.683630  2.680498
#>  [85]  2.680778  2.681055  2.681324  2.683821  2.683857  2.684102  2.684343
#>  [92]  2.684266  2.684506  2.677798  2.678014  2.678229  2.678443  2.678655
#>  [99]  2.678865  2.679074
#> 
#> $fold
#>  [1]  6  5  5  8  9  3  1 10  9  2  1  7 10  4  6  4  7  8  1  3  2
#> 
#> $lambda
#>   [1] 5.773338732 5.384235024 5.021355604 4.682933042 4.367319028 4.072976343
#>   [7] 3.798471369 3.542467112 3.303716685 3.081057237 2.873404291 2.679746459
#>  [13] 2.499140517 2.330706810 2.173624972 2.027129923 1.890508151 1.763094229
#>  [19] 1.644267579 1.533449447 1.430100087 1.333716127 1.243828124 1.159998272
#>  [25] 1.081818270 1.008907340 0.940910362 0.877496152 0.818355849 0.763201403
#>  [31] 0.711764183 0.663793659 0.619056188 0.577333873 0.538423502 0.502135560
#>  [37] 0.468293304 0.436731903 0.407297634 0.379847137 0.354246711 0.330371668
#>  [43] 0.308105724 0.287340429 0.267974646 0.249914052 0.233070681 0.217362497
#>  [49] 0.202712992 0.189050815 0.176309423 0.164426758 0.153344945 0.143010009
#>  [55] 0.133371613 0.124382812 0.115999827 0.108181827 0.100890734 0.094091036
#>  [61] 0.087749615 0.081835585 0.076320140 0.071176418 0.066379366 0.061905619
#>  [67] 0.057733387 0.053842350 0.050213556 0.046829330 0.043673190 0.040729763
#>  [73] 0.037984714 0.035424671 0.033037167 0.030810572 0.028734043 0.026797465
#>  [79] 0.024991405 0.023307068 0.021736250 0.020271299 0.018905082 0.017630942
#>  [85] 0.016442676 0.015334494 0.014301001 0.013337161 0.012438281 0.011599983
#>  [91] 0.010818183 0.010089073 0.009409104 0.008774962 0.008183558 0.007632014
#>  [97] 0.007117642 0.006637937 0.006190562 0.005773339
#> 
#> $fit
#> $beta
#>               5.7733     5.3842     5.0214     4.6829    4.3673    4.0730
#> (Intercept) 20.01905 22.1972413 24.2286321 26.1231139 27.889914 29.537638
#> am           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> cyl          0.00000 -0.3363387 -0.6500094 -0.9425396 -1.215354 -1.469782
#> disp         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> wt           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#>                3.7985    3.5425    3.3037    3.0811    2.8734    2.6797
#> (Intercept) 31.074310 32.507416 33.843936 35.090378 36.252815 37.336907
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.707063 -1.928351 -2.134725 -2.327191 -2.506685 -2.674081
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                2.4991    2.3307    2.1736    2.0271    1.8905    1.7631
#> (Intercept) 38.347935 39.290823 40.170163 40.990239 41.565060 41.565060
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.830196 -2.975789 -3.111569 -3.238199 -3.326958 -3.326958
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.6443    1.5334    1.4301    1.3337    1.2438    1.1600
#> (Intercept) 41.565060 41.565060 41.565060 41.565060 41.565060 41.565060
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -3.326958 -3.326958 -3.326958 -3.326958 -3.326958 -3.326958
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.0818    1.0089    0.9409     0.8775    0.8184    0.7632
#> (Intercept) 41.565060 41.565060 41.565060 35.3604459  8.746550  8.048697
#> am           0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> cyl         -3.326958 -3.326958 -3.326958 -1.6618709  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.2897354  1.539720  1.577546
#> vs           0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000 -2.9314867 -4.745957 -4.734485
#>                0.7118    0.6638    0.6191    0.5773    0.5384    0.5021
#> (Intercept)  8.048618  8.048614  8.048614  8.048614  8.048614  8.048614
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         1.577550  1.577550  1.577550  1.577550  1.577550  1.577550
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -4.734484 -4.734484 -4.734484 -4.734484 -4.734484 -4.734484
#>                0.4683    0.4367    0.4073    0.3798     0.3542     0.3304
#> (Intercept)  1.949718 -1.027366 -1.030217 -1.032134 -15.240293 -15.246078
#> am           1.855900  2.761538  2.762273  2.762767   3.423216   3.424043
#> carb         0.000000  0.000000  0.000000  0.000000   0.000000   0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000   0.000000   0.000000
#> disp         0.000000  0.000000  0.000000  0.000000   0.000000   0.000000
#> drat         0.000000  0.000000  0.000000  0.000000   0.000000   0.000000
#> gear         0.000000  0.000000  0.000000  0.000000   0.000000   0.000000
#> hp           0.000000  0.000000  0.000000  0.000000   0.000000   0.000000
#> qsec         1.744936  1.826669  1.826759  1.826819   2.812842   2.813122
#> vs           0.000000  0.000000  0.000000  0.000000  -4.271370  -4.271979
#> wt          -4.036598 -3.696030 -3.695747 -3.695556  -4.270879  -4.270664
#>                 0.3081     0.2873     0.2680     0.2499     0.2331      0.2174
#> (Intercept) -15.250352 -15.253491 -15.255781 -15.257441 -21.821650 -53.0853339
#> am            3.424667   3.425137   3.425488   3.425749   2.296269   1.8325975
#> carb          0.000000   0.000000   0.000000   0.000000   0.000000   0.0000000
#> cyl           0.000000   0.000000   0.000000   0.000000   0.000000   0.0000000
#> disp          0.000000   0.000000   0.000000   0.000000   0.000000   0.0000000
#> drat          0.000000   0.000000   0.000000   0.000000   1.702409   2.4256606
#> gear          0.000000   0.000000   0.000000   0.000000   0.000000   0.0000000
#> hp            0.000000   0.000000   0.000000   0.000000   0.000000   0.0350542
#> qsec          2.813327   2.813475   2.813582   2.813658   2.844573   4.5739085
#> vs           -4.272412  -4.272715  -4.272926  -4.273069  -4.790644  -7.8606614
#> wt           -4.270496  -4.270366  -4.270265  -4.270188  -4.080031  -5.8403307
#>                   0.2027       0.1891       0.1763       0.1644      0.1533
#> (Intercept) -53.09259573 -53.09948351 -53.10601651 -53.11221303 -53.1180904
#> am            1.83250811   1.83242333   1.83234291   1.83226664   1.8321943
#> carb          0.00000000   0.00000000   0.00000000   0.00000000   0.0000000
#> cyl           0.00000000   0.00000000   0.00000000   0.00000000   0.0000000
#> disp          0.00000000   0.00000000   0.00000000   0.00000000   0.0000000
#> drat          2.42586628   2.42606136   2.42624639   2.42642189   2.4265884
#> gear          0.00000000   0.00000000   0.00000000   0.00000000   0.0000000
#> hp            0.03506094   0.03506733   0.03507339   0.03507914   0.0350846
#> qsec          4.57430790   4.57468670   4.57504599   4.57538678   4.5757100
#> vs           -7.86151601  -7.86232659  -7.86309542  -7.86382465  -7.8645163
#> wt           -5.84068956  -5.84102990  -5.84135270  -5.84165888  -5.8419493
#>                   0.1430       0.1334       0.1244       0.1160       0.1082
#> (Intercept) -53.12366502 -53.12895253 -50.88608919 -50.87891087 -50.87209427
#> am            1.83212567   1.83206058   2.12011831   2.12021722   2.12031112
#> carb          0.00000000   0.00000000  -0.26926692  -0.26927851  -0.26928952
#> cyl           0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> disp          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> drat          2.42674624   2.42689599   2.51389849   2.51370509   2.51352142
#> gear          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> hp            0.03508977   0.03509468   0.03736112   0.03735463   0.03734846
#> qsec          4.57601659   4.57630739   4.39663643   4.39623829   4.39586021
#> vs           -7.86517237  -7.86579463  -7.56756913  -7.56672161  -7.56591679
#> wt           -5.84222474  -5.84248600  -5.60058536  -5.60022268  -5.59987829
#>                  0.1009       0.0941       0.0877       0.0818       0.0763
#> (Intercept) -50.8656212 -50.85947428 -50.85363717 -50.84809423 -50.84283064
#> am            2.1204003   2.12048496   2.12056536   2.12064171   2.12071420
#> carb         -0.2693000  -0.26930988  -0.26931929  -0.26932823  -0.26933672
#> cyl           0.0000000   0.00000000   0.00000000   0.00000000   0.00000000
#> disp          0.0000000   0.00000000   0.00000000   0.00000000   0.00000000
#> drat          2.5133470   2.51318138   2.51302410   2.51287474   2.51273290
#> gear          0.0000000   0.00000000   0.00000000   0.00000000   0.00000000
#> hp            0.0373426   0.03733703   0.03733175   0.03732673   0.03732197
#> qsec          4.3955012   4.39516026   4.39483652   4.39452909   4.39423716
#> vs           -7.5651525  -7.56442681  -7.56373765  -7.56308323  -7.56246179
#> wt           -5.5995513  -5.59924071  -5.59894582  -5.59866579  -5.59839988
#>                   0.0712       0.0664       0.0619       0.0577       0.0538
#> (Intercept) -50.83783234 -50.83308596 -50.82857880 -50.82429882 -50.82023455
#> am            2.12078305   2.12084842   2.12091050   2.12096945   2.12102543
#> carb         -0.26934478  -0.26935244  -0.26935971  -0.26936661  -0.26937316
#> cyl           0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> disp          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> drat          2.51259821   2.51247030   2.51234884   2.51223350   2.51212397
#> gear          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> hp            0.03731744   0.03731315   0.03730907   0.03730519   0.03730152
#> qsec          4.39395994   4.39369669   4.39344671   4.39320934   4.39298392
#> vs           -7.56187167  -7.56131130  -7.56077917  -7.56027386  -7.55979402
#> wt           -5.59814737  -5.59790759  -5.59767990  -5.59746369  -5.59725837
#>                   0.0502      0.0468       0.0437       0.0407       0.0380
#> (Intercept) -50.81637514 -50.8127103 -50.80923009 -50.80592535 -50.80278717
#> am            2.12107859   2.1211291   2.12117701   2.12122253   2.12126576
#> carb         -0.26937938  -0.2693853  -0.26939090  -0.26939623  -0.26940129
#> cyl           0.00000000   0.0000000   0.00000000   0.00000000   0.00000000
#> disp          0.00000000   0.0000000   0.00000000   0.00000000   0.00000000
#> drat          2.51201996   2.5119212   2.51182741   2.51173834   2.51165377
#> gear          0.00000000   0.0000000   0.00000000   0.00000000   0.00000000
#> hp            0.03729802   0.0372947   0.03729155   0.03728856   0.03728572
#> qsec          4.39276987   4.3925666   4.39237359   4.39219030   4.39201626
#> vs           -7.55933837  -7.5589057  -7.55849481  -7.55810464  -7.55773414
#> wt           -5.59706340  -5.5968783  -5.59670245  -5.59653550  -5.59637697
#>                   0.0354       0.0330       0.0308       0.0287       0.0268
#> (Intercept) -50.79980717 -50.79697738 -50.79429022 -50.79173851 -50.78931541
#> am            2.12130681   2.12134579   2.12138281   2.12141796   2.12145134
#> carb         -0.26940610  -0.26941066  -0.26941499  -0.26941911  -0.26942301
#> cyl           0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> disp          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> drat          2.51157346   2.51149720   2.51142478   2.51135601   2.51129070
#> gear          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> hp            0.03728303   0.03728046   0.03727803   0.03727572   0.03727353
#> qsec          4.39185098   4.39169403   4.39154500   4.39140347   4.39126908
#> vs           -7.55738231  -7.55704822  -7.55673096  -7.55642970  -7.55614362
#> wt           -5.59622643  -5.59608347  -5.59594772  -5.59581882  -5.59569641
#>                   0.0250        0.0233        0.0217        0.0203
#> (Intercept) -50.78701445 -51.893204502 -51.900006423 -51.906553145
#> am            2.12148303   2.242374588   2.242445238   2.242509905
#> carb         -0.26942672  -0.143102247  -0.143022973  -0.142948349
#> cyl           0.00000000   0.000000000   0.000000000   0.000000000
#> disp          0.00000000   0.003857686   0.003859616   0.003861426
#> drat          2.51122869   2.449170224   2.449187069   2.449206590
#> gear          0.00000000   0.000000000   0.000000000   0.000000000
#> hp            0.03727145   0.034324862   0.034329415   0.034333841
#> qsec          4.39114147   4.477156580   4.477571538   4.477970157
#> vs           -7.55587197  -7.542067695  -7.542826183  -7.543557491
#> wt           -5.59558017  -5.936037613  -5.936561175  -5.937060837
#>                    0.0189        0.0176        0.0164        0.0153
#> (Intercept) -51.912852498 -51.918912185 -51.924739771 -51.930342682
#> am            2.242568986   2.242622855   2.242671865   2.242716346
#> carb         -0.142878125  -0.142812062  -0.142749935  -0.142691530
#> cyl           0.000000000   0.000000000   0.000000000   0.000000000
#> disp          0.003863122   0.003864712   0.003866199   0.003867592
#> drat          2.449228491   2.449252499   2.449278363   2.449305848
#> gear          0.000000000   0.000000000   0.000000000   0.000000000
#> hp            0.034338139   0.034342311   0.034346359   0.034350285
#> qsec          4.478352987   4.478720566   4.479073420   4.479412059
#> vs           -7.544262372  -7.544941572  -7.545595828  -7.546225873
#> wt           -5.937537605  -5.937992448  -5.938426299  -5.938840054
#>                    0.0143        0.0133        0.0124        0.0116
#> (Intercept) -51.935728201 -51.940903458 -51.945875431 -51.950650943
#> am            2.242756611   2.242792953   2.242825646   2.242854950
#> carb         -0.142636642  -0.142585079  -0.142536657  -0.142491202
#> cyl           0.000000000   0.000000000   0.000000000   0.000000000
#> disp          0.003868895   0.003870113   0.003871251   0.003872314
#> drat          2.449334738   2.449364833   2.449395950   2.449427918
#> gear          0.000000000   0.000000000   0.000000000   0.000000000
#> hp            0.034354090   0.034357777   0.034361347   0.034364802
#> qsec          4.479736983   4.480048678   4.480347616   4.480634256
#> vs           -7.546832427  -7.547416201  -7.547977895  -7.548518199
#> wt           -5.939234575  -5.939610691  -5.939969199  -5.940310863
#>                    0.0108        0.0101        0.0094        0.0088
#> (Intercept) -51.955236660 -51.959639086 -51.963864569 -51.967919293
#> am            2.242881107   2.242904346   2.242924881   2.242942912
#> carb         -0.142448549  -0.142408542  -0.142371032  -0.142335878
#> cyl           0.000000000   0.000000000   0.000000000   0.000000000
#> disp          0.003873306   0.003874232   0.003875095   0.003875899
#> drat          2.449460581   2.449493798   2.449527435   2.449561374
#> gear          0.000000000   0.000000000   0.000000000   0.000000000
#> hp            0.034368146   0.034371380   0.034374506   0.034377528
#> qsec          4.480909047   4.481172421   4.481424800   4.481666593
#> vs           -7.549037789  -7.549537329  -7.550017470  -7.550478848
#> wt           -5.940636418  -5.940946571  -5.941241998  -5.941523351
#>                    0.0082        0.0076        0.0071        0.0066
#> (Intercept) -51.541720173 -51.538696639 -51.535657839 -51.532609759
#> am            2.314365921   2.314198624   2.314033726   2.313872334
#> carb         -0.135688105  -0.135836768  -0.135979771  -0.136117141
#> cyl           0.000000000   0.000000000   0.000000000   0.000000000
#> disp          0.003697199   0.003692708   0.003688411   0.003684303
#> drat          2.482408566   2.482675694   2.482927202   2.483163952
#> gear         -0.095682657  -0.095776270  -0.095862506  -0.095942451
#> hp            0.034535024   0.034536478   0.034537735   0.034538806
#> qsec          4.467067316   4.466834141   4.466602672   4.466373228
#> vs           -7.533831050  -7.533612707  -7.533386074  -7.533152216
#> wt           -5.933397117  -5.932853261  -5.932326154  -5.931815550
#>                    0.0062        0.0058
#> (Intercept) -51.529558161 -51.526508559
#> am            2.313715327   2.313563388
#> carb         -0.136248941  -0.136375258
#> cyl           0.000000000   0.000000000
#> disp          0.003680377   0.003676627
#> drat          2.483386761   2.483596401
#> gear         -0.096017031  -0.096087030
#> hp            0.034539705   0.034540441
#> qsec          4.466146102   4.465921562
#> vs           -7.532912148  -7.532666830
#> wt           -5.931321175  -5.930842733
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1   1   1   1   1 117  53   3   1   1   1   1   1   1
#>  [37] 138  20   1   1  35   1   1   1   1   1  34 110   1   1   1   1   1   1
#>  [55]   1  83   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1  95   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1  44   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.773338732 5.384235024 5.021355604 4.682933042 4.367319028 4.072976343
#>   [7] 3.798471369 3.542467112 3.303716685 3.081057237 2.873404291 2.679746459
#>  [13] 2.499140517 2.330706810 2.173624972 2.027129923 1.890508151 1.763094229
#>  [19] 1.644267579 1.533449447 1.430100087 1.333716127 1.243828124 1.159998272
#>  [25] 1.081818270 1.008907340 0.940910362 0.877496152 0.818355849 0.763201403
#>  [31] 0.711764183 0.663793659 0.619056188 0.577333873 0.538423502 0.502135560
#>  [37] 0.468293304 0.436731903 0.407297634 0.379847137 0.354246711 0.330371668
#>  [43] 0.308105724 0.287340429 0.267974646 0.249914052 0.233070681 0.217362497
#>  [49] 0.202712992 0.189050815 0.176309423 0.164426758 0.153344945 0.143010009
#>  [55] 0.133371613 0.124382812 0.115999827 0.108181827 0.100890734 0.094091036
#>  [61] 0.087749615 0.081835585 0.076320140 0.071176418 0.066379366 0.061905619
#>  [67] 0.057733387 0.053842350 0.050213556 0.046829330 0.043673190 0.040729763
#>  [73] 0.037984714 0.035424671 0.033037167 0.030810572 0.028734043 0.026797465
#>  [79] 0.024991405 0.023307068 0.021736250 0.020271299 0.018905082 0.017630942
#>  [85] 0.016442676 0.015334494 0.014301001 0.013337161 0.012438281 0.011599983
#>  [91] 0.010818183 0.010089073 0.009409104 0.008774962 0.008183558 0.007632014
#>  [97] 0.007117642 0.006637937 0.006190562 0.005773339
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 27
#> 
#> $loss
#>   [1] 890.51238 756.14118 643.71969 550.08917 472.52254 408.66672 356.49273
#>   [8] 314.25224 280.43992 253.76078 233.10177 217.50724 206.15756 198.35064
#>  [15] 193.48589 191.05034 190.55214 190.55214 190.55214 190.55214 190.55214
#>  [22] 190.55214 190.55214 190.55214 190.55214 190.55214 190.55214 121.56071
#>  [29] 102.71573 102.65229 102.65229 102.65229 102.65229 102.65229 102.65229
#>  [36] 102.65229  90.07446  88.55307  88.55306  88.55306  72.69975  72.69974
#>  [43]  72.69974  72.69973  72.69973  72.69973  67.73780  55.34245  55.34243
#>  [50]  55.34241  55.34240  55.34238  55.34237  55.34236  55.34235  54.05872
#>  [57]  54.05870  54.05869  54.05867  54.05866  54.05864  54.05863  54.05862
#>  [64]  54.05861  54.05861  54.05860  54.05859  54.05858  54.05858  54.05857
#>  [71]  54.05857  54.05857  54.05856  54.05856  54.05856  54.05855  54.05855
#>  [78]  54.05855  54.05855  53.79061  53.79059  53.79057  53.79055  53.79053
#>  [85]  53.79052  53.79050  53.79049  53.79048  53.79047  53.79046  53.79045
#>  [92]  53.79044  53.79043  53.79042  53.77735  53.77733  53.77732  53.77730
#>  [99]  53.77728  53.77727
#> 
#> $linear.predictors
#>      5.7733   5.3842   5.0214   4.6829   4.3673   4.0730   3.7985   3.5425
#> 1  20.01905 20.17921 20.32858 20.46788 20.59779 20.71894 20.83193 20.93731
#> 2  20.01905 20.85189 21.62859 22.35296 23.02850 23.65851 24.24606 24.79401
#> 3  20.01905 20.17921 20.32858 20.46788 20.59779 20.71894 20.83193 20.93731
#> 4  20.01905 19.50653 19.02856 18.58280 18.16708 17.77938 17.41781 17.08061
#> 5  20.01905 19.50653 19.02856 18.58280 18.16708 17.77938 17.41781 17.08061
#> 6  20.01905 20.17921 20.32858 20.46788 20.59779 20.71894 20.83193 20.93731
#> 7  20.01905 19.50653 19.02856 18.58280 18.16708 17.77938 17.41781 17.08061
#> 8  20.01905 19.50653 19.02856 18.58280 18.16708 17.77938 17.41781 17.08061
#> 9  20.01905 19.50653 19.02856 18.58280 18.16708 17.77938 17.41781 17.08061
#> 10 20.01905 19.50653 19.02856 18.58280 18.16708 17.77938 17.41781 17.08061
#> 11 20.01905 19.50653 19.02856 18.58280 18.16708 17.77938 17.41781 17.08061
#> 12 20.01905 20.85189 21.62859 22.35296 23.02850 23.65851 24.24606 24.79401
#> 13 20.01905 20.85189 21.62859 22.35296 23.02850 23.65851 24.24606 24.79401
#> 14 20.01905 20.85189 21.62859 22.35296 23.02850 23.65851 24.24606 24.79401
#> 15 20.01905 19.50653 19.02856 18.58280 18.16708 17.77938 17.41781 17.08061
#> 16 20.01905 19.50653 19.02856 18.58280 18.16708 17.77938 17.41781 17.08061
#> 17 20.01905 20.85189 21.62859 22.35296 23.02850 23.65851 24.24606 24.79401
#> 18 20.01905 20.85189 21.62859 22.35296 23.02850 23.65851 24.24606 24.79401
#> 19 20.01905 19.50653 19.02856 18.58280 18.16708 17.77938 17.41781 17.08061
#> 20 20.01905 20.17921 20.32858 20.46788 20.59779 20.71894 20.83193 20.93731
#> 21 20.01905 19.50653 19.02856 18.58280 18.16708 17.77938 17.41781 17.08061
#>      3.3037   3.0811   2.8734   2.6797   2.4991   2.3307   2.1736   2.0271
#> 1  21.03558 21.12723 21.21271 21.29242 21.36676 21.43609 21.50075 21.56105
#> 2  25.30503 25.78162 26.22608 26.64058 27.02715 27.38767 27.72389 28.03744
#> 3  21.03558 21.12723 21.21271 21.29242 21.36676 21.43609 21.50075 21.56105
#> 4  16.76613 16.47285 16.19934 15.94426 15.70637 15.48451 15.27761 15.08465
#> 5  16.76613 16.47285 16.19934 15.94426 15.70637 15.48451 15.27761 15.08465
#> 6  21.03558 21.12723 21.21271 21.29242 21.36676 21.43609 21.50075 21.56105
#> 7  16.76613 16.47285 16.19934 15.94426 15.70637 15.48451 15.27761 15.08465
#> 8  16.76613 16.47285 16.19934 15.94426 15.70637 15.48451 15.27761 15.08465
#> 9  16.76613 16.47285 16.19934 15.94426 15.70637 15.48451 15.27761 15.08465
#> 10 16.76613 16.47285 16.19934 15.94426 15.70637 15.48451 15.27761 15.08465
#> 11 16.76613 16.47285 16.19934 15.94426 15.70637 15.48451 15.27761 15.08465
#> 12 25.30503 25.78162 26.22608 26.64058 27.02715 27.38767 27.72389 28.03744
#> 13 25.30503 25.78162 26.22608 26.64058 27.02715 27.38767 27.72389 28.03744
#> 14 25.30503 25.78162 26.22608 26.64058 27.02715 27.38767 27.72389 28.03744
#> 15 16.76613 16.47285 16.19934 15.94426 15.70637 15.48451 15.27761 15.08465
#> 16 16.76613 16.47285 16.19934 15.94426 15.70637 15.48451 15.27761 15.08465
#> 17 25.30503 25.78162 26.22608 26.64058 27.02715 27.38767 27.72389 28.03744
#> 18 25.30503 25.78162 26.22608 26.64058 27.02715 27.38767 27.72389 28.03744
#> 19 16.76613 16.47285 16.19934 15.94426 15.70637 15.48451 15.27761 15.08465
#> 20 21.03558 21.12723 21.21271 21.29242 21.36676 21.43609 21.50075 21.56105
#> 21 16.76613 16.47285 16.19934 15.94426 15.70637 15.48451 15.27761 15.08465
#>      1.8905   1.7631   1.6443   1.5334   1.4301   1.3337   1.2438   1.1600
#> 1  21.60331 21.60331 21.60331 21.60331 21.60331 21.60331 21.60331 21.60331
#> 2  28.25723 28.25723 28.25723 28.25723 28.25723 28.25723 28.25723 28.25723
#> 3  21.60331 21.60331 21.60331 21.60331 21.60331 21.60331 21.60331 21.60331
#> 4  14.94940 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940
#> 5  14.94940 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940
#> 6  21.60331 21.60331 21.60331 21.60331 21.60331 21.60331 21.60331 21.60331
#> 7  14.94940 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940
#> 8  14.94940 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940
#> 9  14.94940 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940
#> 10 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940
#> 11 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940
#> 12 28.25723 28.25723 28.25723 28.25723 28.25723 28.25723 28.25723 28.25723
#> 13 28.25723 28.25723 28.25723 28.25723 28.25723 28.25723 28.25723 28.25723
#> 14 28.25723 28.25723 28.25723 28.25723 28.25723 28.25723 28.25723 28.25723
#> 15 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940
#> 16 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940
#> 17 28.25723 28.25723 28.25723 28.25723 28.25723 28.25723 28.25723 28.25723
#> 18 28.25723 28.25723 28.25723 28.25723 28.25723 28.25723 28.25723 28.25723
#> 19 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940
#> 20 21.60331 21.60331 21.60331 21.60331 21.60331 21.60331 21.60331 21.60331
#> 21 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940 14.94940
#>      1.0818   1.0089   0.9409   0.8775   0.8184   0.7632   0.7118   0.6638
#> 1  21.60331 21.60331 21.60331 22.47777 21.65593 21.61075 21.61074 21.61074
#> 2  28.25723 28.25723 28.25723 27.30389 26.39012 26.42281 26.42282 26.42282
#> 3  21.60331 21.60331 21.60331 21.59695 23.42046 23.49481 23.49482 23.49482
#> 4  14.94940 14.94940 14.94940 16.91246 18.62649 18.61189 18.61189 18.61189
#> 5  14.94940 14.94940 14.94940 16.18948 16.19265 16.13491 16.13490 16.13490
#> 6  21.60331 21.60331 21.60331 20.78090 21.52117 21.57768 21.57769 21.57769
#> 7  14.94940 14.94940 14.94940 15.17572 16.22163 16.22864 16.22864 16.22864
#> 8  14.94940 14.94940 14.94940 16.23038 18.14320 18.15387 18.15387 18.15387
#> 9  14.94940 14.94940 14.94940 11.88462 11.51444 11.55692 11.55693 11.55693
#> 10 14.94940 14.94940 14.94940 11.32818 10.44229 10.48071 10.48072 10.48072
#> 11 14.94940 14.94940 14.94940 11.44387 10.20133 10.22372 10.22372 10.22372
#> 12 28.25723 28.25723 28.25723 27.90484 28.28379 28.34764 28.34765 28.34765
#> 13 28.25723 28.25723 28.25723 29.34451 29.59744 29.61865 29.61865 29.61865
#> 14 28.25723 28.25723 28.25723 29.09942 30.67815 30.75407 30.75408 30.75408
#> 15 14.94940 14.94940 14.94940 16.63448 18.01586 17.99650 17.99650 17.99650
#> 16 14.94940 14.94940 14.94940 15.73390 16.75057 16.74175 16.74175 16.74175
#> 17 28.25723 28.25723 28.25723 28.51653 28.66383 28.70308 28.70308 28.70308
#> 18 28.25723 28.25723 28.25723 27.27816 24.30353 24.26191 24.26191 24.26191
#> 19 14.94940 14.94940 14.94940 16.97383 16.02781 15.91479 15.91478 15.91478
#> 20 21.60331 21.60331 21.60331 21.75990 19.46591 19.38613 19.38612 19.38612
#> 21 14.94940 14.94940 14.94940 15.83021 14.28340 14.17875 14.17874 14.17874
#>      0.6191   0.5773   0.5384   0.5021   0.4683   0.4367   0.4073   0.3798
#> 1  21.61074 21.61074 21.61074 21.61074 21.95138 22.11754 22.11765 22.11772
#> 2  26.42282 26.42282 26.42282 26.42282 26.91397 27.15369 27.15390 27.15405
#> 3  23.49482 23.49482 23.49482 23.49482 22.89361 22.60034 22.60015 22.60002
#> 4  18.61189 18.61189 18.61189 18.61189 17.76263 17.34819 17.34785 17.34761
#> 5  16.13490 16.13490 16.13490 16.13490 15.17885 14.71224 14.71182 14.71155
#> 6  21.57769 21.57769 21.57769 21.57769 21.04311 20.78233 20.78215 20.78203
#> 7  16.22864 16.22864 16.22864 16.22864 15.88265 15.71383 15.71369 15.71361
#> 8  18.15387 18.15387 18.15387 18.15387 17.60408 17.33581 17.33560 17.33546
#> 9  11.55693 11.55693 11.55693 11.55693 12.13153 12.41198 12.41223 12.41240
#> 10 10.48072 10.48072 10.48072 10.48072 11.14997 11.47660 11.47689 11.47709
#> 11 10.22372 10.22372 10.22372 10.22372 10.77089 11.03792 11.03815 11.03831
#> 12 28.34765 28.34765 28.34765 28.34765 28.89901 29.16815 29.16840 29.16858
#> 13 29.61865 29.61865 29.61865 29.61865 29.60273 29.59499 29.59500 29.59500
#> 14 30.75408 30.75408 30.75408 30.75408 31.12269 31.30266 31.30286 31.30299
#> 15 17.99650 17.99650 17.99650 17.99650 17.17796 16.77851 16.77817 16.77795
#> 16 16.74175 16.74175 16.74175 16.74175 16.18016 15.90610 15.90587 15.90572
#> 17 28.70308 28.70308 28.70308 28.70308 28.97409 29.10639 29.10652 29.10661
#> 18 24.26191 24.26191 24.26191 24.26191 24.30773 24.33004 24.33003 24.33002
#> 19 15.91478 15.91478 15.91478 15.91478 16.31117 16.50445 16.50454 16.50460
#> 20 19.38612 19.38612 19.38612 19.38612 19.67075 19.80953 19.80960 19.80964
#> 21 14.17874 14.17874 14.17874 14.17874 14.87103 15.20871 15.20892 15.20906
#>      0.3542   0.3304   0.3081   0.2873   0.2680   0.2499   0.2331   0.2174
#> 1  23.29261 23.29282 23.29298 23.29309 23.29317 23.29323 23.24600 22.04817
#> 2  26.35011 26.35025 26.35037 26.35046 26.35053 26.35058 25.71008 25.05631
#> 3  21.43912 21.43886 21.43867 21.43853 21.43842 21.43835 20.81232 20.52112
#> 4  17.94246 17.94218 17.94197 17.94180 17.94167 17.94157 17.92026 18.44717
#> 5  14.06809 14.06751 14.06708 14.06675 14.06651 14.06633 14.13541 14.89005
#> 6  18.95923 18.95887 18.95861 18.95842 18.95828 18.95818 19.78827 19.23039
#> 7  16.32069 16.32065 16.32062 16.32059 16.32057 16.32055 16.29459 16.48706
#> 8  18.33535 18.33530 18.33525 18.33521 18.33517 18.33514 18.25071 19.38756
#> 9  12.91250 12.91288 12.91316 12.91337 12.91353 12.91365 12.89167 12.78510
#> 10 11.71931 11.71968 11.71996 11.72017 11.72033 11.72045 11.84578 11.55740
#> 11 10.93157 10.93182 10.93200 10.93214 10.93225 10.93233 11.42183 11.27294
#> 12 29.28166 29.28202 29.28229 29.28249 29.28264 29.28275 29.03757 29.30215
#> 13 29.10793 29.10789 29.10787 29.10785 29.10784 29.10783 30.16909 29.94458
#> 14 32.05005 32.05045 32.05075 32.05097 32.05113 32.05124 31.98828 33.70519
#> 15 17.17886 17.17856 17.17833 17.17815 17.17801 17.17791 16.50323 15.47149
#> 16 16.29714 16.29696 16.29682 16.29671 16.29662 16.29656 16.23402 16.04925
#> 17 28.81012 28.81026 28.81037 28.81045 28.81052 28.81056 28.49737 28.24271
#> 18 26.01771 26.01789 26.01801 26.01810 26.01816 26.01820 26.78939 26.56884
#> 19 15.43045 15.43024 15.43008 15.42998 15.42990 15.42986 15.97139 16.04569
#> 20 19.95165 19.95162 19.95161 19.95160 19.95159 19.95159 19.42653 18.38051
#> 21 14.00338 14.00328 14.00322 14.00318 14.00316 14.00315 13.46620 15.00634
#>      0.2027   0.1891   0.1763   0.1644   0.1533   0.1430   0.1334   0.1244
#> 1  22.04800 22.04783 22.04767 22.04752 22.04738 22.04725 22.04712 21.76599
#> 2  25.05612 25.05594 25.05577 25.05561 25.05546 25.05532 25.05518 25.37833
#> 3  20.52099 20.52086 20.52075 20.52063 20.52053 20.52043 20.52033 20.59434
#> 4  18.44730 18.44742 18.44753 18.44764 18.44775 18.44784 18.44794 18.59709
#> 5  14.89014 14.89023 14.89032 14.89040 14.89047 14.89054 14.89061 14.90856
#> 6  19.23023 19.23007 19.22992 19.22977 19.22964 19.22951 19.22939 18.74959
#> 7  16.48713 16.48720 16.48727 16.48733 16.48738 16.48744 16.48749 16.45587
#> 8  19.38783 19.38809 19.38833 19.38857 19.38879 19.38900 19.38920 19.23940
#> 9  12.78512 12.78514 12.78516 12.78517 12.78519 12.78520 12.78522 12.71005
#> 10 11.55737 11.55735 11.55732 11.55730 11.55728 11.55726 11.55724 11.58167
#> 11 11.27293 11.27292 11.27291 11.27290 11.27290 11.27289 11.27288 11.40407
#> 12 29.30221 29.30227 29.30233 29.30238 29.30244 29.30249 29.30253 29.40096
#> 13 29.94455 29.94453 29.94451 29.94449 29.94447 29.94445 29.94443 29.84499
#> 14 33.70558 33.70594 33.70629 33.70663 33.70694 33.70724 33.70752 33.65031
#> 15 15.47128 15.47109 15.47090 15.47072 15.47055 15.47039 15.47024 15.57510
#> 16 16.04924 16.04922 16.04920 16.04918 16.04917 16.04915 16.04914 16.28478
#> 17 28.24264 28.24257 28.24251 28.24245 28.24240 28.24235 28.24230 28.37903
#> 18 26.56891 26.56898 26.56905 26.56911 26.56917 26.56923 26.56929 26.67050
#> 19 16.04564 16.04559 16.04554 16.04550 16.04546 16.04542 16.04538 16.62632
#> 20 18.38028 18.38006 18.37985 18.37965 18.37946 18.37929 18.37912 17.89118
#> 21 15.00653 15.00670 15.00687 15.00703 15.00717 15.00732 15.00745 14.69187
#>      0.1160   0.1082   0.1009   0.0941   0.0877   0.0818   0.0763   0.0712
#> 1  21.76615 21.76630 21.76644 21.76658 21.76671 21.76683 21.76695 21.76706
#> 2  25.37853 25.37872 25.37889 25.37906 25.37922 25.37937 25.37952 25.37965
#> 3  20.59447 20.59459 20.59471 20.59482 20.59492 20.59502 20.59512 20.59521
#> 4  18.59697 18.59686 18.59675 18.59665 18.59655 18.59646 18.59637 18.59629
#> 5  14.90847 14.90838 14.90830 14.90822 14.90815 14.90807 14.90801 14.90794
#> 6  18.74973 18.74987 18.75000 18.75013 18.75024 18.75036 18.75046 18.75056
#> 7  16.45580 16.45573 16.45567 16.45561 16.45555 16.45549 16.45544 16.45539
#> 8  19.23912 19.23886 19.23862 19.23838 19.23816 19.23795 19.23775 19.23755
#> 9  12.71002 12.71000 12.70999 12.70997 12.70995 12.70993 12.70992 12.70990
#> 10 11.58169 11.58172 11.58174 11.58177 11.58179 11.58181 11.58183 11.58185
#> 11 11.40409 11.40410 11.40412 11.40413 11.40414 11.40415 11.40417 11.40418
#> 12 29.40090 29.40084 29.40079 29.40074 29.40069 29.40065 29.40060 29.40056
#> 13 29.84501 29.84503 29.84505 29.84507 29.84509 29.84510 29.84512 29.84514
#> 14 33.64993 33.64956 33.64922 33.64889 33.64858 33.64828 33.64800 33.64773
#> 15 15.57531 15.57550 15.57569 15.57587 15.57604 15.57620 15.57635 15.57649
#> 16 16.28481 16.28484 16.28486 16.28489 16.28491 16.28493 16.28495 16.28497
#> 17 28.37910 28.37917 28.37923 28.37930 28.37936 28.37941 28.37946 28.37951
#> 18 26.67044 26.67037 26.67031 26.67026 26.67020 26.67015 26.67010 26.67005
#> 19 16.62640 16.62647 16.62654 16.62660 16.62667 16.62672 16.62678 16.62683
#> 20 17.89138 17.89158 17.89176 17.89194 17.89210 17.89226 17.89241 17.89255
#> 21 14.69168 14.69149 14.69132 14.69115 14.69099 14.69084 14.69070 14.69056
#>      0.0664   0.0619   0.0577   0.0538   0.0502   0.0468   0.0437   0.0407
#> 1  21.76716 21.76726 21.76736 21.76745 21.76753 21.76761 21.76769 21.76776
#> 2  25.37978 25.37991 25.38002 25.38013 25.38024 25.38034 25.38043 25.38052
#> 3  20.59530 20.59538 20.59546 20.59553 20.59560 20.59566 20.59573 20.59579
#> 4  18.59621 18.59613 18.59606 18.59599 18.59593 18.59587 18.59581 18.59575
#> 5  14.90788 14.90782 14.90777 14.90772 14.90767 14.90762 14.90757 14.90753
#> 6  18.75066 18.75075 18.75084 18.75092 18.75100 18.75107 18.75114 18.75121
#> 7  16.45535 16.45530 16.45526 16.45522 16.45518 16.45514 16.45511 16.45508
#> 8  19.23737 19.23720 19.23704 19.23688 19.23674 19.23660 19.23646 19.23634
#> 9  12.70989 12.70988 12.70986 12.70985 12.70984 12.70983 12.70982 12.70981
#> 10 11.58187 11.58188 11.58190 11.58191 11.58193 11.58194 11.58196 11.58197
#> 11 11.40419 11.40420 11.40421 11.40422 11.40422 11.40423 11.40424 11.40425
#> 12 29.40052 29.40049 29.40045 29.40042 29.40039 29.40036 29.40033 29.40030
#> 13 29.84515 29.84516 29.84518 29.84519 29.84520 29.84521 29.84522 29.84524
#> 14 33.64748 33.64724 33.64701 33.64680 33.64659 33.64640 33.64621 33.64603
#> 15 15.57663 15.57676 15.57688 15.57700 15.57711 15.57721 15.57731 15.57741
#> 16 16.28499 16.28501 16.28503 16.28504 16.28506 16.28507 16.28509 16.28510
#> 17 28.37956 28.37961 28.37965 28.37969 28.37973 28.37977 28.37980 28.37983
#> 18 26.67001 26.66997 26.66993 26.66989 26.66985 26.66982 26.66979 26.66976
#> 19 16.62688 16.62693 16.62698 16.62702 16.62706 16.62710 16.62714 16.62717
#> 20 17.89269 17.89281 17.89294 17.89305 17.89316 17.89326 17.89336 17.89346
#> 21 14.69043 14.69031 14.69020 14.69009 14.68998 14.68988 14.68979 14.68970
#>      0.0380   0.0354   0.0330   0.0308   0.0287   0.0268   0.0250   0.0233
#> 1  21.76783 21.76790 21.76796 21.76802 21.76808 21.76813 21.76818 21.86307
#> 2  25.38061 25.38069 25.38077 25.38084 25.38091 25.38098 25.38104 25.25042
#> 3  20.59584 20.59590 20.59595 20.59600 20.59604 20.59609 20.59613 20.68765
#> 4  18.59570 18.59565 18.59560 18.59556 18.59552 18.59548 18.59544 18.71233
#> 5  14.90749 14.90745 14.90742 14.90738 14.90735 14.90732 14.90729 14.92109
#> 6  18.75127 18.75133 18.75139 18.75144 18.75149 18.75154 18.75159 18.65986
#> 7  16.45504 16.45502 16.45499 16.45496 16.45494 16.45491 16.45489 16.18172
#> 8  19.23622 19.23610 19.23599 19.23589 19.23579 19.23570 19.23561 19.09540
#> 9  12.70980 12.70979 12.70978 12.70977 12.70977 12.70976 12.70975 12.90296
#> 10 11.58198 11.58199 11.58200 11.58201 11.58202 11.58203 11.58204 11.62214
#> 11 11.40425 11.40426 11.40427 11.40427 11.40428 11.40428 11.40429 11.30125
#> 12 29.40028 29.40025 29.40023 29.40021 29.40019 29.40017 29.40015 29.33661
#> 13 29.84524 29.84525 29.84526 29.84527 29.84528 29.84529 29.84529 30.00247
#> 14 33.64587 33.64571 33.64556 33.64541 33.64528 33.64515 33.64502 33.70768
#> 15 15.57750 15.57759 15.57767 15.57775 15.57782 15.57789 15.57795 15.59055
#> 16 16.28511 16.28512 16.28514 16.28515 16.28516 16.28517 16.28518 16.42542
#> 17 28.37987 28.37990 28.37992 28.37995 28.37998 28.38000 28.38002 28.35884
#> 18 26.66973 26.66970 26.66967 26.66965 26.66962 26.66960 26.66958 26.56583
#> 19 16.62720 16.62724 16.62727 16.62729 16.62732 16.62735 16.62737 16.62960
#> 20 17.89355 17.89363 17.89371 17.89379 17.89386 17.89393 17.89399 17.87587
#> 21 14.68961 14.68953 14.68945 14.68938 14.68931 14.68925 14.68918 14.70924
#>      0.0217   0.0203   0.0189   0.0176   0.0164   0.0153   0.0143   0.0133
#> 1  21.86299 21.86291 21.86283 21.86276 21.86268 21.86261 21.86254 21.86247
#> 2  25.25022 25.25002 25.24983 25.24965 25.24948 25.24931 25.24915 25.24900
#> 3  20.68760 20.68756 20.68751 20.68746 20.68742 20.68737 20.68733 20.68728
#> 4  18.71249 18.71265 18.71280 18.71294 18.71307 18.71320 18.71333 18.71344
#> 5  14.92117 14.92125 14.92133 14.92140 14.92147 14.92154 14.92161 14.92167
#> 6  18.65961 18.65937 18.65914 18.65892 18.65871 18.65852 18.65833 18.65815
#> 7  16.18165 16.18158 16.18152 16.18147 16.18142 16.18138 16.18134 16.18130
#> 8  19.09559 19.09578 19.09596 19.09613 19.09630 19.09646 19.09662 19.09678
#> 9  12.90308 12.90319 12.90330 12.90340 12.90350 12.90359 12.90367 12.90375
#> 10 11.62213 11.62212 11.62210 11.62209 11.62208 11.62207 11.62206 11.62204
#> 11 11.30115 11.30105 11.30096 11.30087 11.30079 11.30071 11.30064 11.30057
#> 12 29.33665 29.33669 29.33672 29.33676 29.33679 29.33683 29.33686 29.33689
#> 13 30.00244 30.00241 30.00239 30.00237 30.00234 30.00232 30.00230 30.00227
#> 14 33.70807 33.70845 33.70881 33.70916 33.70949 33.70981 33.71012 33.71041
#> 15 15.59041 15.59027 15.59014 15.59001 15.58988 15.58976 15.58964 15.58953
#> 16 16.42546 16.42549 16.42552 16.42556 16.42558 16.42561 16.42563 16.42565
#> 17 28.35878 28.35872 28.35867 28.35862 28.35856 28.35852 28.35847 28.35842
#> 18 26.56578 26.56575 26.56571 26.56568 26.56565 26.56563 26.56560 26.56558
#> 19 16.62950 16.62939 16.62930 16.62920 16.62912 16.62903 16.62895 16.62887
#> 20 17.87573 17.87560 17.87547 17.87535 17.87523 17.87511 17.87500 17.87489
#> 21 14.70950 14.70974 14.70998 14.71021 14.71042 14.71063 14.71083 14.71102
#>      0.0124   0.0116   0.0108   0.0101   0.0094   0.0088   0.0082   0.0076
#> 1  21.86240 21.86233 21.86227 21.86220 21.86214 21.86208 21.90139 21.90135
#> 2  25.24885 25.24871 25.24858 25.24845 25.24833 25.24821 25.25537 25.25552
#> 3  20.68724 20.68719 20.68715 20.68711 20.68707 20.68703 20.66418 20.66403
#> 4  18.71356 18.71366 18.71377 18.71386 18.71396 18.71404 18.71266 18.71249
#> 5  14.92174 14.92179 14.92185 14.92191 14.92196 14.92201 14.96520 14.96520
#> 6  18.65798 18.65781 18.65766 18.65751 18.65737 18.65723 18.61415 18.61436
#> 7  16.18127 16.18124 16.18121 16.18119 16.18117 16.18115 16.19920 16.19949
#> 8  19.09693 19.09707 19.09721 19.09735 19.09748 19.09760 19.10997 19.11003
#> 9  12.90383 12.90390 12.90396 12.90402 12.90408 12.90414 12.89423 12.89400
#> 10 11.62203 11.62202 11.62201 11.62200 11.62199 11.62198 11.62184 11.62183
#> 11 11.30051 11.30045 11.30039 11.30034 11.30028 11.30024 11.31879 11.31900
#> 12 29.33692 29.33695 29.33698 29.33701 29.33704 29.33707 29.33923 29.33927
#> 13 30.00225 30.00223 30.00221 30.00219 30.00217 30.00215 30.04633 30.04635
#> 14 33.71069 33.71097 33.71123 33.71148 33.71171 33.71194 33.71066 33.71048
#> 15 15.58942 15.58931 15.58921 15.58911 15.58901 15.58892 15.58113 15.58109
#> 16 16.42567 16.42569 16.42571 16.42572 16.42573 16.42574 16.41777 16.41761
#> 17 28.35838 28.35834 28.35830 28.35826 28.35822 28.35818 28.36646 28.36649
#> 18 26.56557 26.56555 26.56554 26.56553 26.56552 26.56551 26.50994 26.51008
#> 19 16.62880 16.62873 16.62866 16.62859 16.62853 16.62847 16.60581 16.60589
#> 20 17.87478 17.87468 17.87458 17.87449 17.87439 17.87431 17.85018 17.85014
#> 21 14.71120 14.71137 14.71154 14.71170 14.71185 14.71200 14.71550 14.71532
#>      0.0071   0.0066   0.0062   0.0058
#> 1  21.90130 21.90127 21.90123 21.90120
#> 2  25.25567 25.25581 25.25595 25.25609
#> 3  20.66389 20.66377 20.66365 20.66354
#> 4  18.71232 18.71216 18.71201 18.71186
#> 5  14.96519 14.96518 14.96517 14.96516
#> 6  18.61457 18.61477 18.61497 18.61517
#> 7  16.19978 16.20005 16.20030 16.20054
#> 8  19.11009 19.11014 19.11018 19.11021
#> 9  12.89378 12.89357 12.89337 12.89317
#> 10 11.62182 11.62181 11.62180 11.62179
#> 11 11.31921 11.31940 11.31959 11.31976
#> 12 29.33931 29.33934 29.33937 29.33940
#> 13 30.04636 30.04637 30.04639 30.04640
#> 14 33.71029 33.71010 33.70991 33.70972
#> 15 15.58105 15.58102 15.58099 15.58097
#> 16 16.41746 16.41732 16.41718 16.41706
#> 17 28.36652 28.36654 28.36657 28.36660
#> 18 26.51021 26.51033 26.51045 26.51057
#> 19 16.60596 16.60603 16.60609 16.60616
#> 20 17.85010 17.85007 17.85005 17.85003
#> 21 14.71514 14.71496 14.71478 14.71461
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 22.8 21.4 18.7 14.3 17.8 16.4 17.3 10.4 10.4 14.7 32.4 30.4 33.9 15.5
#> [16] 19.2 27.3 26.0 15.8 19.7 15.0
#> 
#> $X
#>               am       carb        cyl       disp        drat       gear
#>  [1,]  1.0488088  0.5640761 -0.2744106 -0.6796472  0.53658145  0.3735437
#>  [2,]  1.0488088 -1.1281521 -1.4269354 -1.0668014  0.45030112  0.3735437
#>  [3,] -0.9534626 -1.1281521 -0.2744106  0.0499897 -0.87841588 -0.9338592
#>  [4,] -0.9534626 -0.5640761  0.8781141  0.8094077 -0.75762342 -0.9338592
#>  [5,] -0.9534626  0.5640761  0.8781141  0.8094077 -0.65408703 -0.9338592
#>  [6,] -0.9534626  0.5640761 -0.2744106 -0.6230631  0.57109358  0.3735437
#>  [7,] -0.9534626  0.0000000  0.8781141  0.1825156 -0.89567194 -0.9338592
#>  [8,] -0.9534626  0.0000000  0.8781141  0.1825156 -0.89567194 -0.9338592
#>  [9,] -0.9534626  0.5640761  0.8781141  1.6432784 -1.13725685 -0.9338592
#> [10,] -0.9534626  0.5640761  0.8781141  1.5539351 -1.01646439 -0.9338592
#> [11,] -0.9534626  0.5640761  0.8781141  1.4050296 -0.61957490 -0.9338592
#> [12,]  1.0488088 -1.1281521 -1.4269354 -1.2849480  0.84719061  0.3735437
#> [13,]  1.0488088 -0.5640761 -1.4269354 -1.3072838  2.31395613  0.3735437
#> [14,]  1.0488088 -1.1281521 -1.4269354 -1.3415321  1.08877552  0.3735437
#> [15,] -0.9534626 -0.5640761  0.8781141  0.4967062 -1.43060995 -0.9338592
#> [16,] -0.9534626 -0.5640761  0.8781141  1.1072186 -0.87841588 -0.9338592
#> [17,]  1.0488088 -1.1281521 -1.4269354 -1.2827144  0.84719061  0.3735437
#> [18,]  1.0488088 -0.5640761 -1.4269354 -0.9752246  1.45115289  1.6809466
#> [19,]  1.0488088  0.5640761  0.8781141  0.7424002  1.08877552  1.6809466
#> [20,]  1.0488088  1.6922282 -0.2744106 -0.7913263  0.05341163  1.6809466
#> [21,]  1.0488088  2.8203804  0.8781141  0.3701365 -0.08463689  1.6809466
#>               hp        qsec         vs          wt
#>  [1,] -0.6429981 -0.66600262 -0.7071068 -0.61201023
#>  [2,] -0.8737019  0.77689653  1.4142136 -0.88733617
#>  [3,] -0.6429981  1.33392272  1.4142136 -0.06594712
#>  [4,]  0.2391048 -0.29017773 -0.7071068  0.14054734
#>  [5,]  1.1890618 -1.08209447 -0.7071068  0.25985524
#>  [6,] -0.4665775  0.97152014  1.4142136  0.14054734
#>  [7,]  0.3069589 -0.03515369 -0.7071068  0.71873181
#>  [8,]  0.3069589  0.09906949 -0.7071068  0.40669574
#>  [9,]  0.6462292  0.35409352 -0.7071068  1.80168050
#> [10,]  0.7819374  0.24671498 -0.7071068  1.96136955
#> [11,]  0.9854996 -0.02173137 -0.7071068  1.88886705
#> [12,] -1.2401139  1.35405620  1.4142136 -0.99746655
#> [13,] -1.4301053  0.71649610  1.4142136 -1.53435213
#> [14,] -1.2536847  1.64263603  1.4142136 -1.33244644
#> [15,] -0.1001655 -0.39084511 -0.7071068  0.21396759
#> [16,]  0.2391048 -0.27004425 -0.7071068  0.51223735
#> [17,] -1.2401139  0.97152014  1.4142136 -1.24067113
#> [18,] -0.9008435 -0.50493481 -0.7071068 -1.05253173
#> [19,]  1.4469072 -1.98138976 -0.7071068 -0.10724601
#> [20,]  0.2391048 -1.31027387 -0.7071068 -0.47434726
#> [21,]  2.4104350 -1.91427817 -0.7071068  0.25985524
#> attr(,"center")
#>  [1]   0.4761905   3.0000000   6.4761905 251.2857143   3.5890476   3.7142857
#>  [7] 157.3809524  17.4523810   0.3333333   3.2868571
#> attr(,"scale")
#>  [1]   0.4994328   1.7728105   1.7353207 134.3133872   0.5795064   0.7648752
#>  [7]  73.6875490   1.4900556   0.4714045   1.0896176
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 57
#> 
#> $lambda.min
#> [1] 0.1159998
#> 
#> $null.dev
#> [1] 42.40535
#> 
#> $Bias
#> [1] 2.903266
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 36.39319