Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 42.87247 39.83375 35.90853 31.96628 27.85244 24.52253 21.85773 19.75581
#>   [9] 18.12892 16.90161 16.00919 15.39625 15.01537 14.82609 14.79386 14.81880
#>  [17] 14.54412 14.42329 14.39354 14.53257 14.74363 14.96154 15.18699 15.32412
#>  [25] 15.42232 15.54442 15.61849 15.31278 15.21699 16.74285 17.75560 18.31009
#>  [33] 18.35342 18.23053 17.97211 16.94739 15.93033 15.02582 14.76419 16.47662
#>  [41] 17.47583 17.62758 17.68456 18.30281 18.30294 18.38745 17.76973 18.50629
#>  [49] 18.73452 18.26471 18.62528 16.65974 18.74564 20.30800 20.30913 19.55571
#>  [57] 19.94642 19.63935 20.80568 21.90325 21.58453 21.30061 21.30160 21.89040
#>  [65] 21.89144 21.71352 24.85042 24.71871 25.25516 25.25461 26.23296 25.76066
#>  [73] 25.76066 25.76071 25.76080 25.76092 25.76107 25.76125 25.74105 25.74147
#>  [81] 25.74189 25.94410 26.20319 26.20375 26.20430 26.20485 26.20539 26.20592
#>  [89] 26.20644 26.20705 26.07976 26.40782 26.40787 26.40793 26.40800 26.40807
#>  [97] 26.40815 26.40823 26.40831 26.40840
#> 
#> $cvse
#>   [1] 11.703189 11.372310 10.755816  9.924578  8.775143  7.803351  6.980698
#>   [8]  6.284830  5.698786  5.210160  4.810152  4.492492  4.252287  4.084877
#>  [15]  3.984922  3.931619  3.800203  3.716173  3.652403  3.632723  3.631640
#>  [22]  3.617974  3.584079  3.536691  3.493396  3.471199  3.466233  3.423178
#>  [29]  3.377263  3.426780  3.523923  3.568281  3.565695  3.453323  3.364963
#>  [36]  3.113040  2.903837  2.740961  2.710344  3.610468  3.783363  3.766323
#>  [43]  3.755611  3.672950  3.672938  3.637578  3.555747  3.545315  3.660669
#>  [50]  3.257515  3.288770  3.419974  4.515505  4.647698  4.647948  4.435894
#>  [57]  4.439994  4.479091  5.396064  5.471655  5.400567  5.375848  5.376527
#>  [64]  5.462632  5.463207  5.445838  7.825370  7.825780  7.777136  7.776938
#>  [71]  7.896994  9.225733  9.225763  9.225797  9.225833  9.225871  9.225912
#>  [78]  9.225955  9.220329  9.220379  9.220430  9.215589  9.251005  9.251070
#>  [85]  9.251136  9.251202  9.251267  9.251333  9.251399  9.251458  9.261182
#>  [92]  9.501899  9.502078  9.502255  9.502431  9.502606  9.502780  9.502953
#>  [99]  9.503124  9.503294
#> 
#> $fold
#>  [1] 10  1  1  7  2  9  5  8 10  2  5  7  3  3  1  6  6  9  4  4  8
#> 
#> $lambda
#>   [1] 5.608661716 5.230656688 4.878127933 4.549358437 4.242746905 3.956799963
#>   [7] 3.690124889 3.441422822 3.209482441 2.993174067 2.791444152 2.603310159
#>  [13] 2.427855767 2.264226414 2.111625132 1.969308666 1.836583853 1.712804248
#>  [19] 1.597366974 1.489709786 1.389308332 1.295673601 1.208349537 1.126910822
#>  [25] 1.050960804 0.980129563 0.914072111 0.852466710 0.795013307 0.741432071
#>  [31] 0.691462031 0.644859804 0.601398412 0.560866172 0.523065669 0.487812793
#>  [37] 0.454935844 0.424274690 0.395679996 0.369012489 0.344142282 0.320948244
#>  [43] 0.299317407 0.279144415 0.260331016 0.242785577 0.226422641 0.211162513
#>  [49] 0.196930867 0.183658385 0.171280425 0.159736697 0.148970979 0.138930833
#>  [55] 0.129567360 0.120834954 0.112691082 0.105096080 0.098012956 0.091407211
#>  [61] 0.085246671 0.079501331 0.074143207 0.069146203 0.064485980 0.060139841
#>  [67] 0.056086617 0.052306567 0.048781279 0.045493584 0.042427469 0.039568000
#>  [73] 0.036901249 0.034414228 0.032094824 0.029931741 0.027914442 0.026033102
#>  [79] 0.024278558 0.022642264 0.021116251 0.019693087 0.018365839 0.017128042
#>  [85] 0.015973670 0.014897098 0.013893083 0.012956736 0.012083495 0.011269108
#>  [91] 0.010509608 0.009801296 0.009140721 0.008524667 0.007950133 0.007414321
#>  [97] 0.006914620 0.006448598 0.006013984 0.005608662
#> 
#> $fit
#> $beta
#>             5.6087     5.2307    4.8781    4.5494    4.2427    3.9568    3.6901
#> (Intercept)   20.8 22.5170474 24.118372 25.611772 27.004522 28.303405 29.514748
#> am             0.0  0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb           0.0  0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl            0.0  0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp           0.0  0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat           0.0  0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear           0.0  0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp             0.0  0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec           0.0  0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs             0.0  0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt             0.0 -0.5287716 -1.021906 -1.481804 -1.910707 -2.310704 -2.683742
#>                3.4414    3.2095   2.9932   2.7914   2.6033    2.4279    2.2642
#> (Intercept) 30.644451 31.698015 32.68057 33.59691 34.45149 35.248471 35.991740
#> am           0.000000  0.000000  0.00000  0.00000  0.00000  0.000000  0.000000
#> carb         0.000000  0.000000  0.00000  0.00000  0.00000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.00000  0.00000  0.00000  0.000000  0.000000
#> disp         0.000000  0.000000  0.00000  0.00000  0.00000  0.000000  0.000000
#> drat         0.000000  0.000000  0.00000  0.00000  0.00000  0.000000  0.000000
#> gear         0.000000  0.000000  0.00000  0.00000  0.00000  0.000000  0.000000
#> hp           0.000000  0.000000  0.00000  0.00000  0.00000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.00000  0.00000  0.00000  0.000000  0.000000
#> vs           0.000000  0.000000  0.00000  0.00000  0.00000  0.000000  0.000000
#> wt          -3.031638 -3.356088 -3.65867 -3.94086 -4.20403 -4.449465 -4.678357
#>                2.1116    1.9693    1.8366    1.7128    1.5974    1.4897
#> (Intercept) 36.684915 37.331372 37.784497 37.784497 37.784497 37.784497
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -4.891823 -5.090902 -5.230444 -5.230444 -5.230444 -5.230444
#>                  1.3893    1.2957    1.2083     1.1269     1.0510     0.9801
#> (Intercept) 36.82167470 35.221915 33.729875 32.3383929 31.0406922 29.8304522
#> am           0.00000000  0.000000  0.000000  0.0000000  0.0000000  0.0000000
#> carb         0.00000000  0.000000  0.000000  0.0000000  0.0000000  0.0000000
#> cyl          0.00000000  0.000000  0.000000  0.0000000  0.0000000  0.0000000
#> disp         0.00000000  0.000000  0.000000  0.0000000  0.0000000  0.0000000
#> drat         0.00000000  0.000000  0.000000  0.0000000  0.0000000  0.0000000
#> gear         0.00000000  0.000000  0.000000  0.0000000  0.0000000  0.0000000
#> hp           0.00000000  0.000000  0.000000  0.0000000  0.0000000  0.0000000
#> qsec         0.05038868  0.134111  0.212196  0.2850183  0.3529326  0.4162697
#> vs           0.00000000  0.000000  0.000000  0.0000000  0.0000000  0.0000000
#> wt          -5.21493691 -5.189171 -5.165141 -5.1427294 -5.1218287 -5.1023366
#>                 0.9141     0.8525     0.7950     0.7414     0.6915     0.6449
#> (Intercept) 28.7017782 27.6491732 26.6675102 25.7520080 24.8982075 24.1019504
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.4753381  0.5304255  0.5818002  0.6297124  0.6743955  0.7160671
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.0841583 -5.0672050 -5.0513944 -5.0366494 -5.0228981 -5.0100736
#>                0.6014    0.5609     0.5231     0.4878     0.4549     0.4243
#> (Intercept) 12.709102 10.038041 10.7265261 12.4017780 13.9654346 14.2791354
#> am           2.684282  3.489360  3.6911796  4.1821804  4.6407940  4.7323723
#> carb         0.000000  0.000000 -0.1244656 -0.4273273 -0.7101195 -0.7667197
#> cyl          0.000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         1.122109  1.204854  1.1560928  1.0374515  0.9266919  0.9045000
#> vs           0.000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -4.080847 -3.814164 -3.6757301 -3.3389033 -3.0243458 -2.9614587
#>                 0.3957     0.3690     0.3441     0.3209    0.2993     0.2791
#> (Intercept) 14.2778261 14.2768845 14.2762078 14.2757217 14.275373 14.2751222
#> am           4.7330875  4.7336011  4.7339699  4.7342347  4.734425  4.7345612
#> carb        -0.7668423 -0.7669302 -0.7669933 -0.7670385 -0.767071 -0.7670943
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> qsec         0.9045194  0.9045334  0.9045435  0.9045507  0.904556  0.9045597
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> wt          -2.9611469 -2.9609230 -2.9607623 -2.9606469 -2.960564 -2.9605047
#>                 0.2603     0.2428     0.2264     0.2112     0.1969     0.1837
#> (Intercept) 14.2749423  8.3668125  8.3656661  8.3646670  8.3637962  8.3630372
#> am           4.7346591  3.4190939  3.4187268  3.4184069  3.4181281  3.4178851
#> carb        -0.7671111 -1.4318883 -1.4321473 -1.4323731 -1.4325699 -1.4327414
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  2.2234359  2.2240805  2.2246424  2.2251320  2.2255588
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.9045624  0.7236662  0.7235839  0.7235122  0.7234497  0.7233952
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.9604620 -1.9752148 -1.9748845 -1.9745967 -1.9743458 -1.9741271
#>                 0.1713       0.1597     0.1490    0.1389     0.1296      0.1208
#> (Intercept)  8.3623757  8.362197981 -9.0854635 -9.090571 -9.0953351 -14.6339795
#> am           3.4176733  3.417140895  3.9442813  3.944368  3.9444496   3.4590024
#> carb        -1.4328909 -1.433013612 -1.9826903 -1.982940 -1.9831725  -2.0188149
#> cyl          0.0000000  0.000000000  1.2357558  1.236069  1.2363615   1.5406781
#> disp         0.0000000  0.000000000  0.0000000  0.000000  0.0000000   0.0000000
#> drat         0.0000000  0.000000000  0.0000000  0.000000  0.0000000   1.7672240
#> gear         2.2259308  2.226426969  3.9633370  3.964008  3.9646331   3.7637550
#> hp           0.0000000  0.000000000  0.0000000  0.000000  0.0000000   0.0000000
#> qsec         0.7233478  0.723165571  0.8629227  0.862949  0.8629736   0.7711942
#> vs           0.0000000  0.001363137  2.4466505  2.447109  2.4475369   2.6923198
#> wt          -1.9739365 -1.973469918 -1.6188573 -1.618650 -1.6184564  -1.6146657
#>                  0.1127      0.1051      0.0980      0.0914      0.0852
#> (Intercept) -14.6388438 -14.6433325 -14.6474754 -14.6512999 -14.6548310
#> am            3.4591715   3.4593260   3.4594672   3.4595964   3.4597146
#> carb         -2.0190920  -2.0193492  -2.0195880  -2.0198096  -2.0200152
#> cyl           1.5409781   1.5412550   1.5415108   1.5417470   1.5419652
#> disp          0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> drat          1.7670108   1.7668132   1.7666300   1.7664604   1.7663034
#> gear          3.7645200   3.7652297   3.7658879   3.7664983   3.7670643
#> hp            0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> qsec          0.7712286   0.7712595   0.7712872   0.7713122   0.7713346
#> vs            2.6927955   2.6932370   2.6936469   2.6940272   2.6943801
#> wt           -1.6144178  -1.6141867  -1.6139715  -1.6137711  -1.6135847
#>                  0.0795      0.0741      0.0691      0.0645     0.0601
#> (Intercept) -14.6580920 -14.6611041 -14.6638866 -14.6664575 -14.668833
#> am            3.4598228   3.4599219   3.4600126   3.4600958   3.460172
#> carb         -2.0202060  -2.0203830  -2.0205471  -2.0206992  -2.020840
#> cyl           1.5421669   1.5423532   1.5425255   1.5426847   1.542832
#> disp          0.0000000   0.0000000   0.0000000   0.0000000   0.000000
#> drat          1.7661580   1.7660235   1.7658991   1.7657840   1.765678
#> gear          3.7675890   3.7680754   3.7685262   3.7689440   3.769331
#> hp            0.0000000   0.0000000   0.0000000   0.0000000   0.000000
#> qsec          0.7713549   0.7713731   0.7713896   0.7714045   0.771418
#> vs            2.6947075   2.6950112   2.6952929   2.6955541   2.695796
#> wt           -1.6134113  -1.6132501  -1.6131003  -1.6129611  -1.612832
#>                  0.0561      0.0523      0.0488      0.0455      0.0424
#> (Intercept) -14.6710292 -14.6730590 -14.6749356 -14.6766707 -14.6782753
#> am            3.4602420   3.4603061   3.4603650   3.4604191   3.4604688
#> carb         -2.0209711  -2.0210923  -2.0212046  -2.0213088  -2.0214053
#> cyl           1.5429679   1.5430938   1.5432102   1.5433178   1.5434174
#> disp          0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> drat          1.7655792   1.7654883   1.7654042   1.7653265   1.7652547
#> gear          3.7696899   3.7700222   3.7703302   3.7706155   3.7708797
#> hp            0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> qsec          0.7714303   0.7714414   0.7714514   0.7714606   0.7714689
#> vs            2.6960208   2.6962290   2.6964220   2.6966008   2.6967666
#> wt           -1.6127118  -1.6126004  -1.6124969  -1.6124010  -1.6123120
#>                  0.0396      0.0369      0.0344      0.0321      0.0299
#> (Intercept) -14.6797594 -14.6811321 -14.6824019 -14.6835768 -14.6846638
#> am            3.4605144   3.4605564   3.4605950   3.4606306   3.4606633
#> carb         -2.0214947  -2.0215775  -2.0216542  -2.0217253  -2.0217912
#> cyl           1.5435095   1.5435948   1.5436736   1.5437466   1.5438142
#> disp          0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> drat          1.7651884   1.7651270   1.7650704   1.7650180   1.7649696
#> gear          3.7711245   3.7713513   3.7715613   3.7717558   3.7719360
#> hp            0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> qsec          0.7714765   0.7714834   0.7714897   0.7714955   0.7715008
#> vs            2.6969202   2.6970626   2.6971945   2.6973167   2.6974300
#> wt           -1.6122295  -1.6121529  -1.6120819  -1.6120161  -1.6119551
#>                    0.0279        0.0260        0.0243        0.0226
#> (Intercept) -14.290516507 -14.295004059 -14.299061913 -1.430590e+01
#> am            3.609653053   3.610376818   3.611023219  3.611599e+00
#> carb         -1.903265374  -1.903170475  -1.903096037 -1.903039e+00
#> cyl           1.615219788   1.615524258   1.615793552  1.616031e+00
#> disp          0.000000000   0.000000000   0.000000000  3.411558e-06
#> drat          1.709404153   1.709084173   1.708790668  1.709330e+00
#> gear          3.693148573   3.693148834   3.693163329  3.693203e+00
#> hp           -0.005076563  -0.005078412  -0.005079801 -5.082453e-03
#> qsec          0.769089639   0.769320247   0.769527072  7.696999e-01
#> vs            2.691987231   2.692041947   2.692092122  2.692205e+00
#> wt           -1.592624895  -1.592831971  -1.593012560 -1.593202e+00
#>                    0.0211        0.0197        0.0184        0.0171
#> (Intercept) -13.579229184 -13.577393991 -13.575585993 -13.573804789
#> am            3.647839618   3.647883714   3.647927155   3.647969952
#> carb         -1.454116509  -1.453832848  -1.453553392  -1.453278077
#> cyl           1.462813959   1.462658792   1.462505925   1.462355324
#> disp          0.009677901   0.009682551   0.009687131   0.009691644
#> drat          2.084506376   2.084604010   2.084700198   2.084794961
#> gear          3.247865007   3.247477424   3.247095585   3.246719405
#> hp           -0.013645849  -0.013650016  -0.013654121  -0.013658165
#> qsec          0.830587911   0.830618261   0.830648161   0.830677618
#> vs            2.772730374   2.772650286   2.772571385   2.772493652
#> wt           -2.475214385  -2.475701646  -2.476181683  -2.476654605
#>                    0.0160        0.0149        0.0139        0.0130
#> (Intercept) -13.572049982 -13.570321183 -13.568618008 -13.566940079
#> am            3.648012114   3.648053651   3.648094573   3.648134887
#> carb         -1.453006843  -1.452739628  -1.452476375  -1.452217023
#> cyl           1.462206955   1.462060786   1.461916783   1.461774915
#> disp          0.009696089   0.009700469   0.009704784   0.009709035
#> drat          2.084888321   2.084980297   2.085070911   2.085160184
#> gear          3.246348799   3.245983685   3.245623982   3.245269611
#> hp           -0.013662150  -0.013666075  -0.013669942  -0.013673752
#> qsec          0.830706638   0.830735228   0.830763393   0.830791142
#> vs            2.772417070   2.772341623   2.772267294   2.772194066
#> wt           -2.477120516  -2.477579520  -2.478031720  -2.478477217
#>                    0.0121        0.0113        0.0105        0.0098
#> (Intercept) -13.565287021 -13.563658467 -13.562054056 -13.560473429
#> am            3.648174604   3.648213732   3.648252281   3.648290257
#> carb         -1.451961515  -1.451709794  -1.451461804  -1.451217489
#> cyl           1.461635151   1.461497458   1.461361807   1.461228167
#> disp          0.009713223   0.009717348   0.009721413   0.009725417
#> drat          2.085248133   2.085334780   2.085420143   2.085504242
#> gear          3.244920491   3.244576546   3.244237699   3.243903874
#> hp           -0.013677505  -0.013681203  -0.013684846  -0.013688435
#> qsec          0.830818479   0.830845410   0.830871943   0.830898082
#> vs            2.772121924   2.772050850   2.771980830   2.771911848
#> wt           -2.478916110  -2.479348497  -2.479774475  -2.480194138
#>                    0.0091        0.0085        0.0080       0.0074
#> (Intercept) -13.558916235 -13.557382127 -13.555870763 -13.55438181
#> am            3.648327671   3.648364530   3.648400843   3.64843662
#> carb         -1.450976796  -1.450739671  -1.450506060  -1.45027591
#> cyl           1.461096509   1.460966802   1.460839019   1.46071313
#> disp          0.009729362   0.009733249   0.009737078   0.00974085
#> drat          2.085587095   2.085668721   2.085749137   2.08582836
#> gear          3.243574997   3.243250995   3.242931795   3.24261733
#> hp           -0.013691971  -0.013695454  -0.013698886  -0.01370227
#> qsec          0.830923833   0.830949203   0.830974197   0.83099882
#> vs            2.771843888   2.771776935   2.771710975   2.77164599
#> wt           -2.480607580  -2.481014894  -2.481416169  -2.48181150
#>                    0.0069        0.0064        0.0060        0.0056
#> (Intercept) -13.552914927 -13.551469795 -13.550046090 -13.548643495
#> am            3.648471863   3.648506585   3.648540793   3.648574494
#> carb         -1.450049175  -1.449825799  -1.449605734  -1.449388931
#> cyl           1.460589107   1.460466924   1.460346552   1.460227965
#> disp          0.009744566   0.009748227   0.009751834   0.009755387
#> drat          2.085906414   2.085983309   2.086059065   2.086133698
#> gear          3.242307519   3.242002305   3.241701614   3.241405381
#> hp           -0.013705597  -0.013708879  -0.013712111  -0.013715296
#> qsec          0.831023079   0.831046978   0.831070523   0.831093719
#> vs            2.771581974   2.771518905   2.771456771   2.771395558
#> wt           -2.482200964  -2.482584659  -2.482962665  -2.483335069
#> 
#> $iter
#>   [1]   0   4   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   4   3   3   3   3   3   3   3   3   3   3   3  49  20  41  50
#>  [37]  50  13   1   1   1   1   1   1   1  50   1   1   1   1   1   2  93   1
#>  [55]   1  74   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1  52   1   1   1 233   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.608661716 5.230656688 4.878127933 4.549358437 4.242746905 3.956799963
#>   [7] 3.690124889 3.441422822 3.209482441 2.993174067 2.791444152 2.603310159
#>  [13] 2.427855767 2.264226414 2.111625132 1.969308666 1.836583853 1.712804248
#>  [19] 1.597366974 1.489709786 1.389308332 1.295673601 1.208349537 1.126910822
#>  [25] 1.050960804 0.980129563 0.914072111 0.852466710 0.795013307 0.741432071
#>  [31] 0.691462031 0.644859804 0.601398412 0.560866172 0.523065669 0.487812793
#>  [37] 0.454935844 0.424274690 0.395679996 0.369012489 0.344142282 0.320948244
#>  [43] 0.299317407 0.279144415 0.260331016 0.242785577 0.226422641 0.211162513
#>  [49] 0.196930867 0.183658385 0.171280425 0.159736697 0.148970979 0.138930833
#>  [55] 0.129567360 0.120834954 0.112691082 0.105096080 0.098012956 0.091407211
#>  [61] 0.085246671 0.079501331 0.074143207 0.069146203 0.064485980 0.060139841
#>  [67] 0.056086617 0.052306567 0.048781279 0.045493584 0.042427469 0.039568000
#>  [73] 0.036901249 0.034414228 0.032094824 0.029931741 0.027914442 0.026033102
#>  [79] 0.024278558 0.022642264 0.021116251 0.019693087 0.018365839 0.017128042
#>  [85] 0.015973670 0.014897098 0.013893083 0.012956736 0.012083495 0.011269108
#>  [91] 0.010509608 0.009801296 0.009140721 0.008524667 0.007950133 0.007414321
#>  [97] 0.006914620 0.006448598 0.006013984 0.005608662
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 32
#> 
#> $loss
#>   [1] 855.40000 728.58500 622.48540 534.12009 460.91532 400.65035 351.41031
#>   [8] 311.54516 279.63424 254.45537 234.95810 220.24051 209.52906 202.16115
#>  [15] 197.56997 195.27137 194.80119 194.80119 194.80119 194.80119 189.54768
#>  [22] 181.53168 174.85759 169.33103 164.78376 161.07076 158.06705 155.66503
#>  [29] 153.77215 152.30886 151.20688 150.40764 124.94194 123.56579 119.03079
#>  [36] 111.34055 108.44012 108.35606 108.35605 108.35605 108.35605 108.35604
#>  [43] 108.35604 108.35604 108.35604  99.13176  99.13175  99.13174  99.13173
#>  [50]  99.13172  99.13172  99.12715  89.80950  89.80948  89.80947  87.61242
#>  [57]  87.61239  87.61238  87.61236  87.61235  87.61233  87.61232  87.61232
#>  [64]  87.61231  87.61230  87.61230  87.61229  87.61229  87.61228  87.61228
#>  [71]  87.61228  87.61228  87.61227  87.61227  87.61227  87.61227  87.40957
#>  [78]  87.40955  87.40954  87.40909  86.76678  86.76675  86.76673  86.76671
#>  [85]  86.76669  86.76666  86.76664  86.76662  86.76660  86.76658  86.76657
#>  [92]  86.76655  86.76653  86.76651  86.76650  86.76648  86.76646  86.76645
#>  [99]  86.76643  86.76642
#> 
#> $linear.predictors
#>    5.6087   5.2307   4.8781   4.5494   4.2427   3.9568   3.6901   3.4414
#> 1    20.8 20.81705 20.83294 20.84777 20.86160 20.87449 20.88652 20.89773
#> 2    20.8 20.69807 20.60302 20.51436 20.43169 20.35458 20.28268 20.21562
#> 3    20.8 20.83027 20.85849 20.88482 20.90937 20.93226 20.95361 20.97353
#> 4    20.8 20.85142 20.89937 20.94409 20.98579 21.02469 21.06096 21.09479
#> 5    20.8 20.69807 20.60302 20.51436 20.43169 20.35458 20.28268 20.21562
#> 6    20.8 20.54473 20.30666 20.08464 19.87758 19.68448 19.50439 19.33644
#> 7    20.8 19.74100 18.75337 17.83230 16.97331 16.17221 15.42510 14.72835
#> 8    20.8 19.64899 18.57555 17.57446 16.64084 15.77015 14.95813 14.20085
#> 9    20.8 19.69076 18.65628 17.69153 16.79179 15.95269 15.17015 14.44035
#> 10   20.8 21.35375 21.87018 22.35180 22.80097 23.21986 23.61052 23.97485
#> 11   20.8 21.54675 22.24317 22.89266 23.49837 24.06326 24.59008 25.08139
#> 12   20.8 21.21363 21.59937 21.95912 22.29463 22.60752 22.89933 23.17146
#> 13   20.8 20.65577 20.52126 20.39582 20.27883 20.16973 20.06798 19.97308
#> 14   20.8 20.70072 20.60813 20.52177 20.44124 20.36614 20.29610 20.23077
#> 15   20.8 20.48392 20.18914 19.91423 19.65785 19.41875 19.19576 18.98780
#> 16   20.8 21.49387 22.14098 22.74448 23.30730 23.83219 24.32171 24.77823
#> 17   20.8 21.38548 21.93149 22.44071 22.91561 23.35850 23.77154 24.15675
#> 18   20.8 21.71702 22.57223 23.36980 24.11362 24.80731 25.45425 26.05758
#> 19   20.8 21.05235 21.28769 21.50717 21.71186 21.90276 22.08078 22.24681
#> 20   20.8 20.62933 20.47017 20.32173 20.18330 20.05419 19.93379 19.82150
#> 21   20.8 21.04706 21.27747 21.49236 21.69276 21.87965 22.05395 22.21650
#>      3.2095   2.9932   2.7914   2.6033   2.4279   2.2642   2.1116    1.9693
#> 1  20.90819 20.91795 20.92705 20.93553 20.94344 20.95082 20.95770 20.964121
#> 2  20.15307 20.09475 20.04035 19.98962 19.94231 19.89819 19.85704 19.818668
#> 3  20.99210 21.00942 21.02557 21.04063 21.05468 21.06778 21.08000 21.091394
#> 4  21.12634 21.15576 21.18320 21.20879 21.23266 21.25491 21.27567 21.295030
#> 5  20.15307 20.09475 20.04035 19.98962 19.94231 19.89819 19.85704 19.818668
#> 6  19.17981 19.03373 18.89750 18.77045 18.65197 18.54147 18.43841 18.342306
#> 7  14.07856 13.47255 12.90740 12.38033 11.88878 11.43036 11.00284 10.604135
#> 8  13.49460 12.83595 12.22169 11.64883 11.11458 10.61633 10.15167  9.718318
#> 9  13.75973 13.12498 12.53301 11.98094 11.46608 10.98592 10.53812 10.120499
#> 10 24.31462 24.63150 24.92702 25.20262 25.45965 25.69935 25.92290 26.131387
#> 11 25.53959 25.96691 26.36543 26.73709 27.08370 27.40695 27.70842 27.989566
#> 12 23.42526 23.66195 23.88269 24.08855 24.28054 24.45959 24.62657 24.782298
#> 13 19.88459 19.80205 19.72508 19.65330 19.58636 19.52392 19.46570 19.411396
#> 14 20.16985 20.11304 20.06006 20.01064 19.96456 19.92158 19.88150 19.844122
#> 15 18.79386 18.61299 18.44430 18.28699 18.14028 18.00346 17.87585 17.756853
#> 16 25.20399 25.60105 25.97135 26.31669 26.63876 26.93912 27.21924 27.480476
#> 17 24.51599 24.85102 25.16347 25.45486 25.72662 25.98006 26.21641 26.436841
#> 18 26.62025 27.14501 27.63439 28.09079 28.51643 28.91339 29.28359 29.628837
#> 19 22.40165 22.54606 22.68073 22.80632 22.92345 23.03269 23.13456 23.229573
#> 20 19.71678 19.61912 19.52804 19.44310 19.36388 19.29000 19.22111 19.156851
#> 21 22.36809 22.50947 22.64132 22.76428 22.87896 22.98591 23.08565 23.178664
#>       1.8366    1.7128    1.5974    1.4897    1.3893    1.2957    1.2083
#> 1  20.968620 20.968620 20.968620 20.968620 21.035209 21.145848 21.249038
#> 2  19.791770 19.791770 19.791770 19.791770 19.739907 19.653736 19.573367
#> 3  21.099381 21.099381 21.099381 21.099381 21.193800 21.350680 21.496996
#> 4  21.308598 21.308598 21.308598 21.308598 21.548524 21.947169 22.318970
#> 5  19.791770 19.791770 19.791770 19.791770 19.804405 19.825398 19.844978
#> 6  18.274941 18.274941 18.274941 18.274941 18.256801 18.226661 18.198550
#> 7  10.324666 10.324666 10.324666 10.324666 10.349244 10.390083 10.428171
#> 8   9.414569  9.414569  9.414569  9.414569  9.433783  9.465709  9.495485
#> 9   9.827774  9.827774  9.827774  9.827774  9.825608  9.822009  9.818653
#> 10 26.277520 26.277520 26.277520 26.277520 26.329881 26.416880 26.498022
#> 11 28.186632 28.186632 28.186632 28.186632 28.255000 28.368596 28.474542
#> 12 24.891453 24.891453 24.891453 24.891453 24.975133 25.114170 25.243845
#> 13 19.373334 19.373334 19.373334 19.373334 19.315154 19.218486 19.128327
#> 14 19.817922 19.817922 19.817922 19.817922 19.780091 19.717233 19.658608
#> 15 17.673440 17.673440 17.673440 17.673440 17.629369 17.556145 17.487851
#> 16 27.663588 27.663588 27.663588 27.663588 27.683118 27.715568 27.745832
#> 17 26.591347 26.591347 26.591347 26.591347 26.503201 26.356743 26.220147
#> 18 29.870835 29.870835 29.870835 29.870835 29.783044 29.637176 29.501129
#> 19 23.296167 23.296167 23.296167 23.296167 23.157324 22.926632 22.711473
#> 20 19.111812 19.111812 19.111812 19.111812 18.940025 18.654595 18.388385
#> 21 23.243863 23.243863 23.243863 23.243863 23.261380 23.290484 23.317630
#>       1.1269    1.0510    0.9801    0.9141    0.8525    0.7950    0.7414
#> 1  21.345273 21.435023 21.518723 21.596782 21.669581 21.737473 21.800789
#> 2  19.498415 19.428514 19.363325 19.302528 19.245830 19.192953 19.143639
#> 3  21.633452 21.760711 21.879392 21.990076 22.093299 22.189566 22.279344
#> 4  22.665714 22.989088 23.290668 23.571922 23.834221 24.078842 24.306976
#> 5  19.863238 19.880268 19.896150 19.910961 19.924774 19.937657 19.949671
#> 6  18.172334 18.147885 18.125083 18.103819 18.083987 18.065492 18.048244
#> 7  10.463692 10.496820 10.527714 10.556527 10.583397 10.608457 10.631827
#> 8   9.523254  9.549152  9.573304  9.595829  9.616835  9.636426  9.654696
#> 9   9.815523  9.812604  9.809881  9.807342  9.804974  9.802766  9.800707
#> 10 26.573694 26.644267 26.710083 26.771463 26.828707 26.882092 26.931880
#> 11 28.573348 28.665495 28.751432 28.831576 28.906319 28.976025 29.041033
#> 12 25.364781 25.477566 25.582749 25.680844 25.772327 25.857645 25.937212
#> 13 19.044244 18.965828 18.892697 18.824495 18.760890 18.701571 18.646250
#> 14 19.603934 19.552944 19.505392 19.461044 19.419685 19.381114 19.345142
#> 15 17.424160 17.364762 17.309366 17.257705 17.209525 17.164592 17.122687
#> 16 27.774057 27.800380 27.824928 27.847822 27.869173 27.889085 27.907656
#> 17 26.092757 25.973953 25.863156 25.759826 25.663460 25.573589 25.489775
#> 18 29.374252 29.255926 29.145575 29.042661 28.946683 28.857174 28.773697
#> 19 22.510816 22.323682 22.149160 21.986401 21.834610 21.693051 21.561031
#> 20 18.140116 17.908580 17.692648 17.491270 17.303463 17.128315 16.964971
#> 21 23.342945 23.366555 23.388573 23.409107 23.428257 23.446117 23.462773
#>       0.6915    0.6449   0.6014   0.5609   0.5231   0.4878   0.4549   0.4243
#> 1  21.859838 21.914907 21.40298 21.19786 21.25903 21.40793 21.54693 21.57481
#> 2  19.097649 19.054759 17.76929 17.42393 17.50978 17.71872 17.91374 17.95287
#> 3  22.363072 22.441157 22.13339 21.96793 21.87387 21.64505 21.43137 21.38864
#> 4  24.519735 24.718154 25.55074 25.61457 25.37357 24.78722 24.23975 24.13015
#> 5  19.960875 19.971324 19.20559 18.96614 18.74065 18.19200 17.67967 17.57719
#> 6  18.032158 18.017156 17.23667 17.01663 16.98989 16.92483 16.86404 16.85194
#> 7  10.653623 10.673950 11.46018 11.67695 11.71763 11.81660 11.90906 11.92751
#> 8   9.671735  9.687626 10.57058 10.82051 10.89308 11.06964 11.23455 11.26750
#> 9   9.798786  9.796995 10.44412 10.63988 10.72102 10.91844 11.10280 11.13965
#> 10 26.978312 27.021614 28.26299 28.59474 28.71576 29.01022 29.28524 29.34019
#> 11 29.101659 29.158200 30.23500 30.50500 30.55452 30.67503 30.78760 30.81006
#> 12 26.011417 26.080621 25.10322 24.74525 24.67480 24.50346 24.34341 24.31147
#> 13 18.594658 18.546543 17.27450 16.93806 17.04231 17.29599 17.53279 17.58028
#> 14 19.311594 19.280308 18.10388 17.78036 17.85187 18.02590 18.18834 18.22094
#> 15 17.083607 17.047161 16.15021 15.91533 16.05580 16.39759 16.71668 16.78061
#> 16 27.924974 27.941125 28.70481 28.91873 29.03086 29.30369 29.55848 29.60942
#> 17 25.411610 25.338713 25.39960 25.48615 25.60946 25.90949 26.18964 26.24570
#> 18 28.695846 28.623242 28.18271 28.11860 28.14536 28.21047 28.27125 28.28343
#> 19 21.437910 21.323086 21.48213 21.63740 21.40858 20.85173 20.33180 20.22770
#> 20 16.812635 16.670567 17.20756 17.50170 17.17858 16.39225 15.65806 15.51104
#> 21 23.478307 23.492793 24.91986 25.33430 25.45357 25.74375 26.01478 26.06891
#>      0.3957   0.3690   0.3441   0.3209   0.2993   0.2791   0.2603   0.2428
#> 1  21.57475 21.57472 21.57469 21.57467 21.57466 21.57465 21.57464 21.32299
#> 2  17.95272 17.95261 17.95253 17.95247 17.95243 17.95240 17.95238 17.69540
#> 3  21.38847 21.38835 21.38826 21.38820 21.38815 21.38812 21.38809 22.56917
#> 4  24.13002 24.12993 24.12987 24.12982 24.12979 24.12976 24.12974 24.74681
#> 5  17.57682 17.57655 17.57636 17.57622 17.57612 17.57605 17.57600 17.98136
#> 6  16.85176 16.85164 16.85155 16.85149 16.85144 16.85141 16.85138 16.11043
#> 7  11.92770 11.92783 11.92792 11.92799 11.92804 11.92808 11.92810 11.95121
#> 8  11.26773 11.26790 11.26803 11.26811 11.26818 11.26822 11.26825 11.49173
#> 9  11.13986 11.14000 11.14011 11.14018 11.14024 11.14028 11.14031 11.35831
#> 10 29.34054 29.34079 29.34097 29.34110 29.34119 29.34126 29.34130 28.99207
#> 11 30.81030 30.81048 30.81060 30.81069 30.81075 30.81080 30.81083 30.02420
#> 12 24.31119 24.31099 24.31085 24.31075 24.31068 24.31062 24.31059 23.21689
#> 13 17.58015 17.58005 17.57999 17.57994 17.57990 17.57988 17.57986 17.42884
#> 14 18.22079 18.22068 18.22061 18.22055 18.22051 18.22048 18.22046 17.90791
#> 15 16.78059 16.78057 16.78056 16.78055 16.78054 16.78054 16.78053 16.91715
#> 16 29.60967 29.60985 29.60998 29.61008 29.61014 29.61019 29.61023 29.10301
#> 17 26.24585 26.24596 26.24604 26.24609 26.24613 26.24616 26.24618 27.89757
#> 18 28.28339 28.28336 28.28334 28.28333 28.28332 28.28331 28.28330 29.28077
#> 19 20.22753 20.22742 20.22733 20.22727 20.22723 20.22719 20.22717 20.05724
#> 20 15.51086 15.51074 15.51065 15.51058 15.51053 15.51050 15.51047 14.96199
#> 21 26.06930 26.06958 26.06978 26.06992 26.07003 26.07010 26.07016 25.78497
#>      0.2264   0.2112   0.1969   0.1837   0.1713   0.1597   0.1490   0.1389
#> 1  21.32298 21.32297 21.32296 21.32296 21.32295 21.32346 22.25363 22.25381
#> 2  17.69541 17.69541 17.69542 17.69542 17.69542 17.69499 17.84329 17.84336
#> 3  22.56949 22.56977 22.57002 22.57023 22.57042 22.57118 22.28648 22.28645
#> 4  24.74688 24.74694 24.74699 24.74704 24.74708 24.74730 24.85371 24.85375
#> 5  17.98138 17.98140 17.98143 17.98144 17.98146 17.98241 18.92092 18.92104
#> 6  16.11022 16.11004 16.10989 16.10975 16.10963 16.10911 15.89162 15.89152
#> 7  11.95121 11.95122 11.95122 11.95123 11.95123 11.95122 11.77618 11.77616
#> 8  11.49181 11.49188 11.49194 11.49199 11.49203 11.49213 11.35643 11.35644
#> 9  11.35839 11.35847 11.35853 11.35858 11.35863 11.35877 11.13915 11.13913
#> 10 28.99200 28.99194 28.99189 28.99184 28.99180 28.99180 29.35877 29.35886
#> 11 30.02397 30.02378 30.02361 30.02346 30.02333 30.02307 30.32071 30.32074
#> 12 23.21658 23.21632 23.21609 23.21589 23.21571 23.21577 21.48813 21.48754
#> 13 17.42888 17.42892 17.42895 17.42898 17.42901 17.42864 17.58434 17.58443
#> 14 17.90789 17.90787 17.90785 17.90784 17.90783 17.90735 18.09300 18.09308
#> 15 16.91729 16.91741 16.91751 16.91760 16.91768 16.91743 17.21354 17.21370
#> 16 29.10290 29.10281 29.10272 29.10265 29.10259 29.10256 29.29590 29.29592
#> 17 27.89810 27.89856 27.89895 27.89930 27.89960 27.89909 26.59960 26.59957
#> 18 29.28107 29.28133 29.28156 29.28176 29.28193 29.28245 30.23386 30.23416
#> 19 20.05703 20.05685 20.05670 20.05656 20.05644 20.05595 19.08496 19.08466
#> 20 14.96160 14.96127 14.96098 14.96072 14.96050 14.96029 15.51938 15.51935
#> 21 25.78490 25.78485 25.78480 25.78475 25.78472 25.78502 25.68640 25.68634
#>      0.1296   0.1208   0.1127   0.1051   0.0980   0.0914   0.0852   0.0795
#> 1  22.25397 21.81877 21.81901 21.81923 21.81944 21.81963 21.81980 21.81997
#> 2  17.84343 18.08311 18.08315 18.08319 18.08323 18.08326 18.08329 18.08332
#> 3  22.28643 22.03260 22.03261 22.03262 22.03263 22.03264 22.03265 22.03267
#> 4  24.85379 24.74011 24.74016 24.74021 24.74025 24.74028 24.74032 24.74035
#> 5  18.92114 19.76809 19.76810 19.76811 19.76812 19.76813 19.76814 19.76815
#> 6  15.89143 15.90196 15.90183 15.90172 15.90161 15.90150 15.90141 15.90132
#> 7  11.77613 11.47449 11.47451 11.47453 11.47454 11.47456 11.47457 11.47458
#> 8  11.35645 11.19386 11.19390 11.19393 11.19397 11.19400 11.19403 11.19406
#> 9  11.13912 11.41940 11.41936 11.41932 11.41928 11.41925 11.41922 11.41919
#> 10 29.35895 29.38942 29.38953 29.38963 29.38972 29.38981 29.38988 29.38995
#> 11 30.32076 30.55780 30.55780 30.55780 30.55780 30.55780 30.55780 30.55779
#> 12 21.48698 21.48368 21.48302 21.48240 21.48184 21.48131 21.48083 21.48038
#> 13 17.58451 17.14904 17.14918 17.14931 17.14943 17.14954 17.14965 17.14974
#> 14 18.09316 18.30712 18.30717 18.30722 18.30726 18.30730 18.30733 18.30737
#> 15 17.21385 17.32860 17.32876 17.32891 17.32905 17.32917 17.32929 17.32940
#> 16 29.29594 29.37773 29.37775 29.37777 29.37779 29.37781 29.37782 29.37784
#> 17 26.59954 27.02124 27.02118 27.02112 27.02107 27.02102 27.02098 27.02094
#> 18 30.23445 29.71383 29.71423 29.71461 29.71496 29.71528 29.71558 29.71586
#> 19 19.08438 18.65321 18.65293 18.65266 18.65242 18.65219 18.65198 18.65178
#> 20 15.51932 15.56976 15.56970 15.56965 15.56960 15.56955 15.56951 15.56946
#> 21 25.68628 25.81618 25.81612 25.81606 25.81601 25.81596 25.81592 25.81588
#>      0.0741   0.0691   0.0645   0.0601   0.0561   0.0523   0.0488   0.0455
#> 1  21.82012 21.82026 21.82039 21.82051 21.82062 21.82072 21.82082 21.82090
#> 2  18.08335 18.08337 18.08339 18.08341 18.08343 18.08345 18.08346 18.08348
#> 3  22.03268 22.03269 22.03270 22.03271 22.03272 22.03273 22.03274 22.03275
#> 4  24.74037 24.74040 24.74042 24.74044 24.74046 24.74048 24.74050 24.74051
#> 5  19.76816 19.76817 19.76817 19.76818 19.76819 19.76820 19.76821 19.76821
#> 6  15.90124 15.90116 15.90109 15.90102 15.90096 15.90090 15.90084 15.90079
#> 7  11.47459 11.47460 11.47461 11.47462 11.47463 11.47463 11.47464 11.47465
#> 8  11.19409 11.19411 11.19413 11.19416 11.19418 11.19419 11.19421 11.19423
#> 9  11.41917 11.41915 11.41913 11.41911 11.41909 11.41908 11.41906 11.41905
#> 10 29.39002 29.39008 29.39014 29.39019 29.39024 29.39028 29.39032 29.39036
#> 11 30.55779 30.55779 30.55778 30.55778 30.55777 30.55777 30.55777 30.55776
#> 12 21.47997 21.47958 21.47922 21.47890 21.47859 21.47831 21.47805 21.47780
#> 13 17.14983 17.14991 17.14999 17.15006 17.15012 17.15018 17.15024 17.15029
#> 14 18.30740 18.30742 18.30745 18.30747 18.30749 18.30751 18.30753 18.30755
#> 15 17.32950 17.32959 17.32968 17.32976 17.32984 17.32990 17.32997 17.33003
#> 16 29.37785 29.37786 29.37787 29.37788 29.37789 29.37790 29.37791 29.37791
#> 17 27.02090 27.02087 27.02084 27.02081 27.02079 27.02077 27.02075 27.02073
#> 18 29.71612 29.71636 29.71658 29.71679 29.71698 29.71716 29.71733 29.71748
#> 19 18.65160 18.65143 18.65128 18.65113 18.65100 18.65087 18.65076 18.65065
#> 20 15.56943 15.56939 15.56935 15.56932 15.56929 15.56926 15.56923 15.56921
#> 21 25.81584 25.81580 25.81577 25.81574 25.81572 25.81569 25.81567 25.81565
#>      0.0424   0.0396   0.0369   0.0344   0.0321   0.0299   0.0279   0.0260
#> 1  21.82099 21.82106 21.82113 21.82120 21.82126 21.82131 21.80633 21.80644
#> 2  18.08349 18.08350 18.08352 18.08353 18.08354 18.08355 18.01166 18.01168
#> 3  22.03276 22.03277 22.03278 22.03279 22.03279 22.03280 22.12269 22.12232
#> 4  24.74052 24.74053 24.74055 24.74056 24.74057 24.74057 24.64239 24.64256
#> 5  19.76822 19.76823 19.76823 19.76824 19.76824 19.76825 19.92448 19.92428
#> 6  15.90075 15.90071 15.90066 15.90063 15.90059 15.90056 15.93047 15.93067
#> 7  11.47465 11.47466 11.47466 11.47467 11.47467 11.47467 11.53244 11.53251
#> 8  11.19424 11.19426 11.19427 11.19428 11.19429 11.19430 11.20116 11.20111
#> 9  11.41904 11.41902 11.41901 11.41900 11.41900 11.41899 11.33635 11.33613
#> 10 29.39040 29.39043 29.39046 29.39048 29.39051 29.39053 29.45105 29.45126
#> 11 30.55776 30.55775 30.55775 30.55775 30.55774 30.55774 30.60746 30.60780
#> 12 21.47758 21.47737 21.47718 21.47700 21.47684 21.47669 21.33456 21.33418
#> 13 17.15034 17.15038 17.15042 17.15046 17.15049 17.15053 17.22913 17.22927
#> 14 18.30757 18.30758 18.30759 18.30760 18.30762 18.30763 18.36188 18.36201
#> 15 17.33008 17.33013 17.33018 17.33022 17.33026 17.33030 17.27006 17.27002
#> 16 29.37792 29.37793 29.37793 29.37794 29.37794 29.37795 29.43471 29.43485
#> 17 27.02071 27.02070 27.02068 27.02067 27.02066 27.02065 26.98550 26.98497
#> 18 29.71762 29.71776 29.71788 29.71799 29.71810 29.71819 29.58999 29.58986
#> 19 18.65055 18.65046 18.65037 18.65029 18.65022 18.65015 18.86557 18.86572
#> 20 15.56918 15.56916 15.56914 15.56912 15.56910 15.56909 15.37419 15.37450
#> 21 25.81563 25.81561 25.81559 25.81558 25.81556 25.81555 25.78794 25.78784
#>      0.0243   0.0226   0.0211   0.0197   0.0184   0.0171   0.0160   0.0149
#> 1  21.80655 21.80643 21.86481 21.86482 21.86483 21.86484 21.86485 21.86486
#> 2  18.01170 18.01174 18.24273 18.24282 18.24291 18.24300 18.24308 18.24317
#> 3  22.12199 22.12166 22.10934 22.10933 22.10932 22.10931 22.10930 22.10928
#> 4  24.64272 24.64292 24.58908 24.58903 24.58899 24.58894 24.58890 24.58885
#> 5  19.92409 19.92404 19.94524 19.94518 19.94512 19.94506 19.94500 19.94494
#> 6  15.93085 15.93065 15.50267 15.50250 15.50234 15.50217 15.50201 15.50185
#> 7  11.53257 11.53273 11.86768 11.86786 11.86804 11.86821 11.86838 11.86855
#> 8  11.20108 11.20115 11.19742 11.19742 11.19742 11.19742 11.19742 11.19742
#> 9  11.33594 11.33598 11.14192 11.14181 11.14171 11.14160 11.14150 11.14140
#> 10 29.45145 29.45171 29.32183 29.32175 29.32167 29.32159 29.32151 29.32144
#> 11 30.60811 30.60856 30.81436 30.81445 30.81454 30.81463 30.81472 30.81481
#> 12 21.33384 21.33338 21.40424 21.40442 21.40459 21.40476 21.40493 21.40510
#> 13 17.22939 17.22911 17.04184 17.04176 17.04168 17.04160 17.04152 17.04144
#> 14 18.36213 18.36209 18.28685 18.28680 18.28675 18.28669 18.28664 18.28659
#> 15 17.27000 17.27007 17.50639 17.50646 17.50653 17.50660 17.50667 17.50674
#> 16 29.43497 29.43518 29.50723 29.50726 29.50729 29.50732 29.50736 29.50739
#> 17 26.98450 26.98453 26.98166 26.98162 26.98159 26.98156 26.98152 26.98149
#> 18 29.58975 29.58956 29.55260 29.55252 29.55245 29.55237 29.55230 29.55223
#> 19 18.86584 18.86568 18.93907 18.93921 18.93934 18.93946 18.93959 18.93971
#> 20 15.37478 15.37497 15.38842 15.38845 15.38847 15.38849 15.38851 15.38853
#> 21 25.78775 25.78785 25.59461 25.59453 25.59444 25.59436 25.59428 25.59420
#>      0.0139   0.0130   0.0121   0.0113   0.0105   0.0098   0.0091   0.0085
#> 1  21.86487 21.86488 21.86489 21.86490 21.86491 21.86492 21.86493 21.86494
#> 2  18.24326 18.24334 18.24342 18.24350 18.24358 18.24366 18.24374 18.24381
#> 3  22.10927 22.10926 22.10925 22.10924 22.10923 22.10922 22.10921 22.10920
#> 4  24.58881 24.58877 24.58872 24.58868 24.58864 24.58860 24.58856 24.58852
#> 5  19.94488 19.94482 19.94476 19.94471 19.94465 19.94460 19.94455 19.94449
#> 6  15.50169 15.50154 15.50139 15.50124 15.50109 15.50094 15.50080 15.50066
#> 7  11.86872 11.86888 11.86905 11.86921 11.86936 11.86952 11.86967 11.86982
#> 8  11.19742 11.19742 11.19742 11.19742 11.19742 11.19742 11.19742 11.19741
#> 9  11.14130 11.14120 11.14110 11.14101 11.14092 11.14082 11.14073 11.14064
#> 10 29.32136 29.32129 29.32122 29.32115 29.32108 29.32101 29.32094 29.32087
#> 11 30.81490 30.81498 30.81507 30.81515 30.81523 30.81531 30.81539 30.81547
#> 12 21.40526 21.40542 21.40558 21.40574 21.40589 21.40605 21.40620 21.40635
#> 13 17.04137 17.04129 17.04122 17.04115 17.04108 17.04101 17.04094 17.04087
#> 14 18.28654 18.28649 18.28644 18.28639 18.28635 18.28630 18.28625 18.28621
#> 15 17.50681 17.50688 17.50695 17.50701 17.50707 17.50714 17.50720 17.50726
#> 16 29.50742 29.50745 29.50748 29.50751 29.50754 29.50757 29.50759 29.50762
#> 17 26.98146 26.98142 26.98139 26.98136 26.98133 26.98130 26.98127 26.98124
#> 18 29.55216 29.55209 29.55202 29.55195 29.55188 29.55181 29.55175 29.55168
#> 19 18.93983 18.93995 18.94007 18.94019 18.94030 18.94042 18.94053 18.94064
#> 20 15.38855 15.38857 15.38859 15.38861 15.38863 15.38865 15.38867 15.38869
#> 21 25.59412 25.59404 25.59397 25.59389 25.59382 25.59375 25.59367 25.59360
#>      0.0080   0.0074   0.0069   0.0064   0.0060   0.0056
#> 1  21.86495 21.86495 21.86496 21.86497 21.86498 21.86499
#> 2  18.24389 18.24396 18.24403 18.24410 18.24418 18.24424
#> 3  22.10919 22.10918 22.10917 22.10916 22.10915 22.10914
#> 4  24.58848 24.58844 24.58841 24.58837 24.58833 24.58830
#> 5  19.94444 19.94439 19.94434 19.94429 19.94424 19.94419
#> 6  15.50052 15.50038 15.50025 15.50011 15.49998 15.49985
#> 7  11.86997 11.87011 11.87026 11.87040 11.87054 11.87068
#> 8  11.19741 11.19741 11.19741 11.19741 11.19741 11.19741
#> 9  11.14055 11.14047 11.14038 11.14030 11.14021 11.14013
#> 10 29.32081 29.32074 29.32068 29.32062 29.32055 29.32049
#> 11 30.81555 30.81562 30.81570 30.81577 30.81584 30.81591
#> 12 21.40649 21.40663 21.40678 21.40692 21.40705 21.40719
#> 13 17.04080 17.04073 17.04067 17.04060 17.04054 17.04048
#> 14 18.28616 18.28612 18.28607 18.28603 18.28599 18.28595
#> 15 17.50732 17.50738 17.50744 17.50750 17.50756 17.50761
#> 16 29.50765 29.50768 29.50770 29.50773 29.50776 29.50778
#> 17 26.98121 26.98118 26.98116 26.98113 26.98110 26.98107
#> 18 29.55162 29.55156 29.55150 29.55144 29.55138 29.55132
#> 19 18.94075 18.94085 18.94096 18.94106 18.94116 18.94126
#> 20 15.38871 15.38872 15.38874 15.38876 15.38878 15.38879
#> 21 25.59353 25.59346 25.59340 25.59333 25.59326 25.59320
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.4 18.7 24.4 22.8 19.2 17.3 10.4 10.4 14.7 32.4 33.9 21.5 15.5 15.2 19.2
#> [16] 27.3 26.0 30.4 19.7 15.0 21.4
#> 
#> $X
#>               am       carb       cyl       disp        drat       gear
#>  [1,] -0.7844645 -0.9497694  0.000000  0.1941551 -0.97364915 -0.9338592
#>  [2,] -0.7844645 -0.3799077  1.080123  0.9589221 -0.82834314 -0.9338592
#>  [3,] -0.7844645 -0.3799077 -1.080123 -0.6403406  0.29258898  0.3735437
#>  [4,] -0.7844645 -0.3799077 -1.080123 -0.6845771  0.77002304  0.3735437
#>  [5,] -0.7844645  0.7598155  0.000000 -0.4836383  0.77002304  0.3735437
#>  [6,] -0.7844645  0.1899539  1.080123  0.3276144 -0.99440716 -0.9338592
#>  [7,] -0.7844645  0.7598155  1.080123  1.7986662 -1.28501919 -0.9338592
#>  [8,] -0.7844645  0.7598155  1.080123  1.7086936 -1.13971317 -0.9338592
#>  [9,] -0.7844645  0.7598155  1.080123  1.5587393 -0.66227912 -0.9338592
#> [10,]  1.2747549 -0.9497694 -1.080123 -1.1501852  1.10215107  0.3735437
#> [11,]  1.2747549 -0.9497694 -1.080123 -1.2071678  1.39276311  0.3735437
#> [12,] -0.7844645 -0.9497694 -1.080123 -0.8397798  0.31334699 -0.9338592
#> [13,] -0.7844645 -0.3799077  1.080123  0.6440180 -1.63790523 -0.9338592
#> [14,] -0.7844645 -0.3799077  1.080123  0.5390500 -0.82834314 -0.9338592
#> [15,] -0.7844645 -0.3799077  1.080123  1.2588307 -0.97364915 -0.9338592
#> [16,]  1.2747549 -0.9497694 -1.080123 -1.1479359  1.10215107  0.3735437
#> [17,]  1.2747549 -0.3799077 -1.080123 -0.8382803  1.82868115  1.6809466
#> [18,]  1.2747549 -0.3799077 -1.080123 -1.0272227  0.45865300  1.6809466
#> [19,]  1.2747549  1.8995387  0.000000 -0.6530867  0.14728297  1.6809466
#> [20,]  1.2747549  3.0392620  1.080123  0.5165569 -0.01878105  1.6809466
#> [21,]  1.2747549 -0.3799077 -1.080123 -0.8330319  1.16442508  0.3735437
#>               hp        qsec         vs          wt
#>  [1,] -0.4840216  0.75230440  1.0488088 -0.03006413
#>  [2,]  0.4918860 -0.61508149 -0.9534626  0.17976309
#>  [3,] -1.2046918  1.06872428  1.0488088 -0.05337827
#>  [4,] -0.7092310  2.70732720  1.0488088 -0.09068089
#>  [5,] -0.2888401  0.10816394  1.0488088  0.17976309
#>  [6,]  0.5669559 -0.28736091 -0.9534626  0.45020707
#>  [7,]  0.9423049 -0.07264742 -0.9534626  1.86770654
#>  [8,]  1.0924446 -0.16305310 -0.9534626  2.02997292
#>  [9,]  1.3176540 -0.38906730 -0.9534626  1.95630025
#> [10,] -1.1446360  0.76925546  1.0488088 -0.97661806
#> [11,] -1.1596499  1.01222073  1.0488088 -1.31700444
#> [12,] -0.6792031  1.07437463  1.0488088 -0.72948821
#> [13,]  0.1165370 -0.69983682 -0.9534626  0.25436833
#> [14,]  0.1165370 -0.45687156 -0.9534626  0.17510026
#> [15,]  0.4918860 -0.59813043 -0.9534626  0.55745209
#> [16,] -1.1446360  0.44718523  1.0488088 -1.22374790
#> [17,] -0.7692869 -0.79589285 -0.9534626 -1.03257198
#> [18,] -0.4389797 -0.68288575  1.0488088 -1.61729051
#> [19,]  0.4918860 -1.47393544 -0.9534626 -0.44505575
#> [20,]  2.8941202 -1.98246739 -0.9534626  0.30099660
#> [21,] -0.4990355  0.27767459  1.0488088 -0.43573010
#> attr(,"center")
#>  [1]   0.3809524   2.6666667   6.0000000 232.1047619   3.5490476   3.7142857
#>  [7] 142.2380952  18.1085714   0.4761905   3.2472381
#> attr(,"scale")
#>  [1]   0.4856209   1.7548120   1.8516402 133.3739657   0.4817419   0.7648752
#>  [7]  66.6046650   1.7698003   0.4994328   1.0723108
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 19
#> 
#> $lambda.min
#> [1] 1.597367
#> 
#> $null.dev
#> [1] 40.73333
#> 
#> $Bias
#> [1] 9.443062
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 9.810706