Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 39.59664 37.52603 33.67952 29.70966 26.51605 23.98012 21.99991 20.48785
#>   [9] 19.36879 18.57833 18.06135 17.77073 17.66627 17.71373 17.88400 18.12893
#>  [17] 17.74847 17.38490 17.30560 17.39621 17.49022 17.51547 17.54022 17.57503
#>  [25] 17.39498 17.54195 15.80644 15.01791 14.64428 12.69004 11.40351 10.23326
#>  [33] 11.36991 11.52306 11.67799 11.89100 12.13629 12.22338 12.20896 12.18190
#>  [41] 13.58945 12.96754 12.82144 12.68638 12.30656 12.28111 12.09811 13.36740
#>  [49] 17.62383 16.95496 16.97648 16.97710 18.79890 17.39338 16.71754 16.59824
#>  [57] 14.84038 19.93307 19.93307 19.84026 19.84036 21.05592 21.05744 21.47758
#>  [65] 21.47836 21.13025 21.13093 21.13157 21.13219 21.13278 21.13335 21.13389
#>  [73] 21.30085 21.30157 21.30227 21.26883 21.26946 21.23122 21.23180 21.23237
#>  [81] 21.23291 21.23345 21.23396 21.23447 21.23495 21.23543 21.23589 21.19357
#>  [89] 21.19403 21.19449 21.19493 21.19536 21.19579 21.19620 21.19661 21.19701
#>  [97] 21.19739 21.19777 21.19814 21.19851
#> 
#> $cvse
#>   [1] 12.848629 12.726546 11.872238 10.670215  9.612394  8.673539  7.836501
#>   [8]  7.090962  6.432320  5.860650  5.379582  4.994820  4.712065  4.534450
#>  [15]  4.460163  4.475180  4.296739  4.181422  4.155865  4.179124  4.198700
#>  [22]  4.210784  4.227079  4.247969  4.238493  4.318881  4.495375  4.550848
#>  [29]  4.568699  3.842531  3.479491  3.324132  4.211676  4.180696  4.151439
#>  [36]  4.115275  4.078358  4.052551  4.035589  4.028193  4.270658  4.120759
#>  [43]  4.107677  4.089746  4.113089  4.114769  4.107564  4.498951  5.347517
#>  [50]  5.392115  5.432205  5.432668  5.664272  5.339670  5.004995  5.018481
#>  [57]  4.334271  6.490861  6.491150  6.503438  6.503725  6.876484  6.877239
#>  [64]  6.870266  6.870982  6.906562  6.907237  6.907891  6.908526  6.909140
#>  [71]  6.909736  6.910313  6.914580  6.915116  6.915635  6.916094  6.916585
#>  [78]  6.920833  6.921296  6.921745  6.922180  6.922603  6.923014  6.923412
#>  [85]  6.923799  6.924174  6.924539  6.928615  6.928957  6.929290  6.929614
#>  [92]  6.929929  6.930235  6.930532  6.930822  6.931104  6.931378  6.931645
#>  [99]  6.931905  6.932158
#> 
#> $fold
#>  [1]  9  9  6  6  4  2  4  8  3  1  2  8  5 10  5 10  7  1  1  7  3
#> 
#> $lambda
#>   [1] 5.265612946 4.910728257 4.579761608 4.271101003 3.983243091 3.714785838
#>   [7] 3.464421705 3.230931277 3.013177323 2.810099256 2.620707971 2.444081025
#>  [13] 2.279358144 2.125737034 1.982469472 1.848857665 1.724250846 1.608042110
#>  [19] 1.499665454 1.398593022 1.304332533 1.216424886 1.134441919 1.057984331
#>  [25] 0.986679728 0.920180817 0.858163709 0.800326347 0.746387030 0.696083042
#>  [31] 0.649169375 0.605417532 0.564614417 0.526561295 0.491072826 0.457976161
#>  [37] 0.427110100 0.398324309 0.371478584 0.346442171 0.323093128 0.301317732
#>  [43] 0.281009926 0.262070797 0.244408103 0.227935814 0.212573703 0.198246947
#>  [49] 0.184885767 0.172425085 0.160804211 0.149966545 0.139859302 0.130433253
#>  [55] 0.121642489 0.113444192 0.105798433 0.098667973 0.092018082 0.085816371
#>  [61] 0.080032635 0.074638703 0.069608304 0.064916938 0.060541753 0.056461442
#>  [67] 0.052656129 0.049107283 0.045797616 0.042711010 0.039832431 0.037147858
#>  [73] 0.034644217 0.032309313 0.030131773 0.028100993 0.026207080 0.024440810
#>  [79] 0.022793581 0.021257370 0.019824695 0.018488577 0.017242508 0.016080421
#>  [85] 0.014996655 0.013985930 0.013043325 0.012164249 0.011344419 0.010579843
#>  [91] 0.009866797 0.009201808 0.008581637 0.008003263 0.007463870 0.006960830
#>  [97] 0.006491694 0.006054175 0.005646144 0.005265613
#> 
#> $fit
#> $beta
#>               5.2656     4.9107    4.5798    4.2711    3.9832    3.7148
#> (Intercept) 20.12857 21.9318062 23.613509 25.181871 26.644530 28.008611
#> am           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000 -0.5523166 -1.067409 -1.547786 -1.995787 -2.413594
#>                3.4644   3.2309    3.0132    2.8101    2.6207    2.4441
#> (Intercept) 29.280757 30.46717 31.573614 32.605491 33.567823 34.465297
#> am           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.803242 -3.16663 -3.505526 -3.821582 -4.116337 -4.391226
#>                2.2794    2.1257    1.9825    1.8489    1.7243    1.6080
#> (Intercept) 35.302285 36.082862 36.810831 37.489737 37.965607 37.965607
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -4.647589 -4.886673 -5.109645 -5.317588 -5.463343 -5.463343
#>                1.4997    1.3986       1.3043      1.2164     1.1344     1.0580
#> (Intercept) 37.965607 37.965607 37.777528210 36.29251109 34.9074929 33.6158204
#> am           0.000000  0.000000  0.000000000  0.00000000  0.0000000  0.0000000
#> carb         0.000000  0.000000  0.000000000  0.00000000  0.0000000  0.0000000
#> cyl          0.000000  0.000000  0.000000000  0.00000000  0.0000000  0.0000000
#> disp         0.000000  0.000000  0.000000000  0.00000000  0.0000000  0.0000000
#> drat         0.000000  0.000000  0.000000000  0.00000000  0.0000000  0.0000000
#> gear         0.000000  0.000000  0.000000000  0.00000000  0.0000000  0.0000000
#> hp           0.000000  0.000000  0.000000000  0.00000000  0.0000000  0.0000000
#> qsec         0.000000  0.000000  0.009853646  0.08765517  0.1602176  0.2278897
#> vs           0.000000  0.000000  0.000000000  0.00000000  0.0000000  0.0000000
#> wt          -5.463343 -5.463343 -5.460381050 -5.43699105 -5.4151761 -5.3948314
#>                 0.9867     0.9202     0.8582    0.8003    0.7464    0.6961
#> (Intercept) 32.4112022 31.2877713 23.8079433 11.603974  9.705594  8.445699
#> am           0.0000000  0.0000000  1.7911004  4.916791  5.191527  5.305768
#> carb         0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> cyl          0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> disp         0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> drat         0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> gear         0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> hp           0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> qsec         0.2910008  0.3498585  0.6045868  1.004626  1.083047  1.140307
#> vs           0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> wt          -5.3758579 -5.3581631 -4.6626487 -3.462267 -3.343753 -3.287061
#>                0.6492    0.6054    0.5646    0.5266    0.4911    0.4580
#> (Intercept)  8.072391  8.070335  8.069110  8.068379  8.067944  8.067684
#> am           5.339919  5.340540  5.340911  5.341132  5.341263  5.341342
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         1.157250  1.157310  1.157346  1.157367  1.157380  1.157387
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -3.270166 -3.269933 -3.269795 -3.269712 -3.269663 -3.269633
#>                 0.4271     0.3983     0.3715    0.3464     0.3231     0.3013
#> (Intercept)  8.8997546 10.2892324 11.5856264 12.794655 13.4756971 13.4749188
#> am           5.4028223  5.5072205  5.6046721  5.695557  5.7456158  5.7462769
#> carb        -0.1111589 -0.2974508 -0.4712806 -0.633396 -0.7243191 -0.7244769
#> cyl          0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> qsec         1.1123880  1.0371306  0.9669117  0.901425  0.8646117  0.8646066
#> vs           0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> wt          -3.1871817 -3.0484828 -2.9190480 -2.798336 -2.7309751 -2.7306421
#>                 0.2810     0.2621     0.2444     0.2279     0.2126     0.1982
#> (Intercept) 13.4743747 13.0038493 11.7954251 11.0857350 11.0868774 11.0876871
#> am           5.7467099  5.6204896  5.2870357  5.0915932  5.0908861  5.0904027
#> carb        -0.7245763 -0.7685303 -0.8831152 -0.9505222 -0.9506578 -0.9507439
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.2768778  0.9973816  1.4203787  1.4208513  1.4211565
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.8646049  0.8331404  0.7508129  0.7024716  0.7023554  0.7022786
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.7304262 -2.6628882 -2.4888615 -2.3864924 -2.3865251 -2.3865569
#>                 0.1849     0.1724     0.1608     0.1500       0.1399
#> (Intercept) 11.0882561 11.0886529 11.0889277 11.0891166 -35.60528244
#> am           5.0900744  5.0898530  5.0897046  5.0896059   6.39141684
#> carb        -0.9507980 -0.9508314 -0.9508517 -0.9508638  -2.17204286
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000   2.61551349
#> disp         0.0000000  0.0000000  0.0000000  0.0000000   0.00000000
#> drat         1.4213517  1.4214751  1.4215521  1.4215995   3.26844665
#> gear         0.0000000  0.0000000  0.0000000  0.0000000   3.18899815
#> hp           0.0000000  0.0000000  0.0000000  0.0000000   0.02277045
#> qsec         0.7022282  0.7021955  0.7021744  0.7021610   1.38311917
#> vs           0.0000000  0.0000000  0.0000000  0.0000000   2.49129343
#> wt          -2.3865850 -2.3866083 -2.3866269 -2.3866413  -2.74935703
#>                   0.1304       0.1216       0.1134       0.1058       0.0987
#> (Intercept) -35.61087443 -35.61616671 -35.62116962 -35.62589352 -35.63034879
#> am            6.39170585   6.39198487   6.39225250   6.39250778   6.39275007
#> carb         -2.17218617  -2.17232134  -2.17244912  -2.17257014  -2.17268491
#> cyl           2.61584643   2.61615588   2.61644340   2.61671047   2.61695849
#> disp          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> drat          3.26864094   3.26882250   3.26899210   3.26915046   3.26929827
#> gear          3.18929979   3.18958171   3.18984606   3.19009465   3.19032896
#> hp            0.02277344   0.02277644   0.02277942   0.02278237   0.02278527
#> qsec          1.38322035   1.38331790   1.38341153   1.38350101   1.38358621
#> vs            2.49155280   2.49179415   2.49201859   2.49222717   2.49242089
#> wt           -2.74944957  -2.74953809  -2.74962214  -2.74970142  -2.74977576
#>                  0.0920       0.0858       0.0800       0.0746       0.0696
#> (Intercept) -35.6345459 -35.63849529 -35.64220754 -35.64569314 -35.64896256
#> am            6.3929790   6.39319442   6.39339643   6.39358524   6.39376120
#> carb         -2.1727939  -2.17289734  -2.17299565  -2.17308903  -2.17317770
#> cyl           2.6171888   2.61740257   2.61760102   2.61778522   2.61795618
#> disp          0.0000000   0.00000000   0.00000000   0.00000000   0.00000000
#> drat          3.2694362   3.26956483   3.26968477   3.26979657   3.26990076
#> gear          3.1905502   3.19075947   3.19095756   3.19114523   3.19132308
#> hp            0.0227881   0.02279084   0.02279349   0.02279604   0.02279847
#> qsec          1.3836670   1.38374349   1.38381559   1.38388339   1.38394701
#> vs            2.4926007   2.49276765   2.49292249   2.49306609   2.49319924
#> wt           -2.7498451  -2.74990951  -2.74996906  -2.75002392  -2.75007428
#>                  0.0649       0.0605       0.0565       0.0527       0.0491
#> (Intercept) -35.6520262 -35.65489430 -35.65757698 -35.66008410 -35.66242531
#> am            6.3939248   6.39407643   6.39421680   6.39434645   6.39446602
#> carb         -2.1732619  -2.17334171  -2.17341737  -2.17348902  -2.17355681
#> cyl           2.6181149   2.61826214   2.61839885   2.61852575   2.61864355
#> disp          0.0000000   0.00000000   0.00000000   0.00000000   0.00000000
#> drat          3.2699978   3.27008820   3.27017236   3.27025071   3.27032363
#> gear          3.1914917   3.19165143   3.19180280   3.19194614   3.19208180
#> hp            0.0228008   0.02280301   0.02280511   0.02280708   0.02280895
#> qsec          1.3840066   1.38406222   1.38411412   1.38416244   1.38420736
#> vs            2.4933227   2.49343711   2.49354317   2.49364149   2.49373262
#> wt           -2.7501204  -2.75016247  -2.75020080  -2.75023562  -2.75026721
#>                  0.0458       0.0427       0.0398        0.0371        0.0346
#> (Intercept) -35.6646100 -35.66664726 -35.66854582 -37.178200492 -37.173545448
#> am            6.3945761   6.39467740   6.39477045   6.456035017   6.456015038
#> carb         -2.1736209  -2.17368136  -2.17373840  -1.898275135  -1.898020367
#> cyl           2.6187529   2.61885448   2.61894879   2.613636259   2.613303201
#> disp          0.0000000   0.00000000   0.00000000   0.009519504   0.009521741
#> drat          3.2703915   3.27045462   3.27051336   3.447095990   3.446835575
#> gear          3.1922101   3.19233132   3.19244579   3.100316249   3.099870206
#> hp            0.0228107   0.02281235   0.02281388   0.014466074   0.014463806
#> qsec          1.3842491   1.38428775   1.38432359   1.494342174   1.494334871
#> vs            2.4938171   2.49389541   2.49396803   2.564103218   2.563877078
#> wt           -2.7502958  -2.75032168  -2.75034505  -3.503736294  -3.503954138
#>                    0.0323        0.0301        0.0281        0.0262
#> (Intercept) -37.169070005 -37.164769528 -37.160639350 -37.156674786
#> am            6.455991474   6.455964713   6.455935117   6.455903023
#> carb         -1.897780160  -1.897553808  -1.897340637  -1.897139999
#> cyl           2.612985167   2.612681624   2.612392047   2.612115925
#> disp          0.009523782   0.009525638   0.009527321   0.009528843
#> drat          3.446586882   3.446349498   3.446123019   3.445907049
#> gear          3.099446203   3.099043337   3.098660736   3.098297549
#> hp            0.014461670   0.014459659   0.014457767   0.014455989
#> qsec          1.494325622   1.494314631   1.494302087   1.494288166
#> vs            2.563660191   2.563452285   2.563253088   2.563062332
#> wt           -3.504152548  -3.504332788  -3.504496055  -3.504643479
#>                    0.0244        0.0228        0.0213        0.0198
#> (Intercept) -37.152871149 -37.149223757 -37.145727950 -37.142379094
#> am            6.455868744   6.455832572   6.455794775   6.455755604
#> carb         -1.896951270  -1.896773856  -1.896607183  -1.896450705
#> cyl           2.611852755   2.611602049   2.611363328   2.611136126
#> disp          0.009530214   0.009531446   0.009532548   0.009533528
#> drat          3.445701198   3.445505088   3.445318347   3.445140613
#> gear          3.097952957   3.097626163   3.097316399   3.097022920
#> hp            0.014454317   0.014452746   0.014451272   0.014449889
#> qsec          1.494273030   1.494256832   1.494239714   1.494221805
#> vs            2.562879748   2.562705073   2.562538046   2.562378409
#> wt           -3.504776132  -3.504895025  -3.505001113  -3.505095298
#>                    0.0185        0.0172        0.0161        0.0150
#> (Intercept) -37.139172596 -37.136103905 -37.133168529 -37.130362030
#> am            6.455715289   6.455674043   6.455632062   6.455589527
#> carb         -1.896303897  -1.896166258  -1.896037307  -1.895916586
#> cyl           2.610919991   2.610714479   2.610519163   2.610333626
#> disp          0.009534395   0.009535159   0.009535825   0.009536402
#> drat          3.444971534   3.444810766   3.444657975   3.444512836
#> gear          3.096745008   3.096481969   3.096233133   3.095997855
#> hp            0.014448593   0.014447378   0.014446240   0.014445176
#> qsec          1.494203227   1.494184092   1.494164502   1.494144552
#> vs            2.562225910   2.562080298   2.561941331   2.561808767
#> wt           -3.505178432  -3.505251316  -3.505314706  -3.505369313
#>                    0.0140        0.0130        0.0122        0.0113
#> (Intercept) -37.127680040 -37.125118263 -37.122672474 -37.120338533
#> am            6.455546602   6.455503440   6.455460178   6.455416942
#> carb         -1.895803656  -1.895698098  -1.895599512  -1.895507517
#> cyl           2.610157463   2.609990281   2.609831701   2.609681353
#> disp          0.009536896   0.009537313   0.009537661   0.009537943
#> drat          3.444375034   3.444244262   3.444120223   3.444002631
#> gear          3.095775512   3.095565505   3.095367256   3.095180210
#> hp            0.014444180   0.014443250   0.014442382   0.014441572
#> qsec          1.494124330   1.494103915   1.494083380   1.494062793
#> vs            2.561682373   2.561561918   2.561447179   2.561337936
#> wt           -3.505415808  -3.505454819  -3.505486939  -3.505512723
#>                    0.0106        0.0099        0.0092        0.0086
#> (Intercept) -37.118112381 -37.115990046 -37.113967643 -37.112041381
#> am            6.455373846   6.455330995   6.455288482   6.455246390
#> carb         -1.895421749  -1.895341860  -1.895267520  -1.895198413
#> cyl           2.609538882   2.609403943   2.609276202   2.609155339
#> disp          0.009538166   0.009538334   0.009538453   0.009538526
#> drat          3.443891206   3.443785680   3.443685791   3.443591288
#> gear          3.095003833   3.094837611   3.094681052   3.094533681
#> hp            0.014440817   0.014440114   0.014439460   0.014438852
#> qsec          1.494042214   1.494021699   1.494001296   1.493981053
#> vs            2.561233976   2.561135090   2.561041077   2.560951738
#> wt           -3.505532692  -3.505547335  -3.505557109  -3.505562444
#>                    0.0080        0.0075        0.0070       0.0065
#> (Intercept) -37.110207557 -37.108462567 -37.106802898 -37.10522513
#> am            6.455204796   6.455163765   6.455123358   6.45508363
#> carb         -1.895134240  -1.895074716  -1.895019567  -1.89496854
#> cyl           2.609041043   2.608933013   2.608830963   2.60873461
#> disp          0.009538557   0.009538551   0.009538511   0.00953844
#> drat          3.443501928   3.443417477   3.443337708   3.44326241
#> gear          3.094395043   3.094264704   3.094142245   3.09402727
#> hp            0.014438287   0.014437763   0.014437278   0.01443683
#> qsec          1.493961008   1.493941198   1.493921656   1.49390241
#> vs            2.560866884   2.560786329   2.560709891   2.56063740
#> wt           -3.505563740  -3.505561371  -3.505555688  -3.50554702
#>                    0.0061        0.0056        0.0053
#> (Intercept) -37.103725949 -37.102302123 -37.100950523
#> am            6.455044615   6.455006365   6.454968909
#> carb         -1.894921381  -1.894877864  -1.894837767
#> cyl           2.608643694   2.608557951   2.608477135
#> disp          0.009538341   0.009538217   0.009538071
#> drat          3.443191357   3.443124362   3.443061227
#> gear          3.093919382   3.093818227   3.093723450
#> hp            0.014436413   0.014436030   0.014435677
#> qsec          1.493883484   1.493864903   1.493846683
#> vs            2.560568678   2.560503569   2.560441913
#> wt           -3.505535663  -3.505521907  -3.505506014
#> 
#> $iter
#>   [1]   0   5   3   3   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   4   3   3   3   3   3  63  67  12  10   7   1   1   1   1   1
#>  [37]  23  25  25  25  11   1   1  43  55  20   1   1   1   1   1   1 200   1
#>  [55]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1 112
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.265612946 4.910728257 4.579761608 4.271101003 3.983243091 3.714785838
#>   [7] 3.464421705 3.230931277 3.013177323 2.810099256 2.620707971 2.444081025
#>  [13] 2.279358144 2.125737034 1.982469472 1.848857665 1.724250846 1.608042110
#>  [19] 1.499665454 1.398593022 1.304332533 1.216424886 1.134441919 1.057984331
#>  [25] 0.986679728 0.920180817 0.858163709 0.800326347 0.746387030 0.696083042
#>  [31] 0.649169375 0.605417532 0.564614417 0.526561295 0.491072826 0.457976161
#>  [37] 0.427110100 0.398324309 0.371478584 0.346442171 0.323093128 0.301317732
#>  [43] 0.281009926 0.262070797 0.244408103 0.227935814 0.212573703 0.198246947
#>  [49] 0.184885767 0.172425085 0.160804211 0.149966545 0.139859302 0.130433253
#>  [55] 0.121642489 0.113444192 0.105798433 0.098667973 0.092018082 0.085816371
#>  [61] 0.080032635 0.074638703 0.069608304 0.064916938 0.060541753 0.056461442
#>  [67] 0.052656129 0.049107283 0.045797616 0.042711010 0.039832431 0.037147858
#>  [73] 0.034644217 0.032309313 0.030131773 0.028100993 0.026207080 0.024440810
#>  [79] 0.022793581 0.021257370 0.019824695 0.018488577 0.017242508 0.016080421
#>  [85] 0.014996655 0.013985930 0.013043325 0.012164249 0.011344419 0.010579843
#>  [91] 0.009866797 0.009201808 0.008581637 0.008003263 0.007463870 0.006960830
#>  [97] 0.006491694 0.006054175 0.005646144 0.005265613
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 26
#> 
#> $loss
#>   [1] 776.14286 664.36649 570.84894 492.96262 428.43899 375.32068 331.91987
#>   [8] 296.78222 268.65552 246.46254 229.27740 216.30512 206.86392 200.36975
#>  [15] 196.32302 194.29701 193.88258 193.88258 193.88258 193.88258 192.92731
#>  [22] 185.81943 179.88558 174.95649 170.88567 167.54675 130.22516 100.94620
#>  [29]  99.94457  99.65986  99.64397  99.64397  99.64397  99.64396  99.64396
#>  [36]  99.64396  97.00495  93.56517  91.46596  90.47432  90.32628  90.32627
#>  [43]  90.32627  88.92807  86.70360  86.34940  86.34939  86.34939  86.34939
#>  [50]  86.34939  86.34939  86.34939  61.97834  61.97833  61.97832  61.97831
#>  [57]  61.97831  61.97830  61.97830  61.97829  61.97829  61.97828  61.97828
#>  [64]  61.97828  61.97827  61.97827  61.97827  61.97827  61.97827  61.97826
#>  [71]  61.97826  61.11129  61.11128  61.11126  61.11125  61.11124  61.11123
#>  [78]  61.11122  61.11122  61.11121  61.11120  61.11120  61.11119  61.11119
#>  [85]  61.11119  61.11118  61.11118  61.11118  61.11118  61.11117  61.11117
#>  [92]  61.11117  61.11117  61.11117  61.11117  61.11116  61.11116  61.11116
#>  [99]  61.11116  61.11116
#> 
#> $linear.predictors
#>      5.2656   4.9107   4.5798   4.2711   3.9832   3.7148   3.4644   3.2309
#> 1  20.12857 20.34390 20.54471 20.73199 20.90664 21.06953 21.22144 21.36310
#> 2  20.12857 20.15611 20.18179 20.20574 20.22808 20.24891 20.26833 20.28645
#> 3  20.12857 20.02079 19.92027 19.82653 19.73911 19.65758 19.58154 19.51063
#> 4  20.12857 19.96004 19.80286 19.65628 19.51957 19.39208 19.27318 19.16230
#> 5  20.12857 20.19201 20.25117 20.30635 20.35780 20.40579 20.45054 20.49228
#> 6  20.12857 20.03184 19.94162 19.85749 19.77902 19.70585 19.63760 19.57396
#> 7  20.12857 20.03184 19.94162 19.85749 19.77902 19.70585 19.63760 19.57396
#> 8  20.12857 19.68388 19.26915 18.88238 18.52168 18.18528 17.87156 17.57898
#> 9  20.12857 19.87167 19.63207 19.40863 19.20025 19.00591 18.82466 18.65564
#> 10 20.12857 19.84405 19.57870 19.33124 19.10046 18.88523 18.68450 18.49730
#> 11 20.12857 18.93604 17.82388 16.78668 15.81938 14.91728 14.07597 13.29137
#> 12 20.12857 18.97967 17.90821 16.90896 15.97705 15.10795 14.29743 13.54153
#> 13 20.12857 20.71671 21.26521 21.77674 22.25380 22.69870 23.11362 23.50058
#> 14 20.12857 20.91831 21.65481 22.34168 22.98226 23.57967 24.13681 24.65640
#> 15 20.12857 20.57035 20.98235 21.36658 21.72492 22.05910 22.37076 22.66142
#> 16 20.12857 19.98765 19.85623 19.73366 19.61936 19.51276 19.41334 19.32063
#> 17 20.12857 19.81091 19.51466 19.23837 18.98071 18.74041 18.51631 18.30731
#> 18 20.12857 20.74985 21.32925 21.86961 22.37355 22.84352 23.28182 23.69058
#> 19 20.12857 21.09615 21.99852 22.84007 23.62490 24.35684 25.03945 25.67605
#> 20 20.12857 20.40189 20.65679 20.89450 21.11620 21.32296 21.51578 21.69560
#> 21 20.12857 20.39637 20.64611 20.87903 21.09624 21.29882 21.48774 21.66393
#>      3.0132   2.8101   2.6207   2.4441   2.2794    2.1257    1.9825    1.8489
#> 1  21.49523 21.61844 21.73335 21.84052 21.94047 22.033676 22.120603 22.201671
#> 2  20.30335 20.31910 20.33380 20.34751 20.36029 20.372207 20.383324 20.393691
#> 3  19.44449 19.38282 19.32530 19.27166 19.22163 19.174972 19.131461 19.090882
#> 4  19.05889 18.96244 18.87250 18.78862 18.71039 18.637438 18.569400 18.505947
#> 5  20.53121 20.56751 20.60136 20.63294 20.66238 20.689841 20.715451 20.739334
#> 6  19.51460 19.45925 19.40762 19.35948 19.31458 19.272705 19.233654 19.197234
#> 7  19.51460 19.45925 19.40762 19.35948 19.31458 19.272705 19.233654 19.197234
#> 8  17.30612 17.05165 16.81433 16.59301 16.38660 16.194101 16.014578 15.847153
#> 9  18.49800 18.35099 18.21389 18.08602 17.96678 17.855570 17.751857 17.655133
#> 10 18.32272 18.15991 18.00807 17.86646 17.73440 17.611237 17.496375 17.389254
#> 11 12.55964 11.87723 11.24081 10.64729 10.09376  9.577546  9.096119  8.647139
#> 12 12.83658 12.17913 11.56600 10.99419 10.46092  9.963593  9.499781  9.067228
#> 13 23.86146 24.19801 24.51188 24.80460 25.07759 25.332180 25.569613 25.791043
#> 14 25.14097 25.59289 26.01434 26.40740 26.77396 27.115816 27.434633 27.731963
#> 15 22.93249 23.18529 23.42105 23.64092 23.84598 24.037212 24.215557 24.381882
#> 16 19.23416 19.15352 19.07832 19.00818 18.94277 18.881772 18.824882 18.771827
#> 17 18.11239 17.93062 17.76109 17.60299 17.45554 17.318036 17.189796 17.070199
#> 18 24.07179 24.42731 24.75886 25.06807 25.35644 25.625381 25.876192 26.110098
#> 19 26.26975 26.82344 27.33981 27.82137 28.27048 28.689325 29.079939 29.444226
#> 20 21.86331 22.01971 22.16557 22.30160 22.42846 22.546777 22.657116 22.760018
#> 21 21.82825 21.98149 22.12441 22.25769 22.38199 22.497910 22.606019 22.706842
#>       1.7243    1.6080    1.4997    1.3986    1.3043    1.2164    1.1344
#> 1  22.258495 22.258495 22.258495 22.258495 22.246642 22.153053 22.065766
#> 2  20.400958 20.400958 20.400958 20.400958 20.413958 20.516601 20.612333
#> 3  19.062439 19.062439 19.062439 19.062439 19.083850 19.252910 19.410584
#> 4  18.461471 18.461471 18.461471 18.461471 18.440050 18.270911 18.113162
#> 5  20.756075 20.756075 20.756075 20.756075 20.802976 21.173293 21.518672
#> 6  19.171706 19.171706 19.171706 19.171706 19.174139 19.193351 19.211270
#> 7  19.171706 19.171706 19.171706 19.171706 19.180051 19.245945 19.307401
#> 8  15.729799 15.729799 15.729799 15.729799 15.725231 15.689157 15.655513
#> 9  17.587336 17.587336 17.587336 17.587336 17.583731 17.555265 17.528717
#> 10 17.314169 17.314169 17.314169 17.314169 17.314653 17.318478 17.322045
#> 11  8.332433  8.332433  8.332433  8.332433  8.336013  8.364287  8.390656
#> 12  8.764037  8.764037  8.764037  8.764037  8.763442  8.758747  8.754368
#> 13 25.946252 25.946252 25.946252 25.946252 25.956540 26.037777 26.113543
#> 14 27.940372 27.940372 27.940372 27.940372 27.953817 28.059970 28.158976
#> 15 24.498466 24.498466 24.498466 24.498466 24.514860 24.644308 24.765039
#> 16 18.734638 18.734638 18.734638 18.734638 18.723218 18.633045 18.548945
#> 17 16.986368 16.986368 16.986368 16.986368 16.961510 16.765232 16.582171
#> 18 26.274052 26.274052 26.274052 26.274052 26.256869 26.121192 25.994651
#> 19 29.699569 29.699569 29.699569 29.699569 29.682498 29.547716 29.422010
#> 20 22.832146 22.832146 22.832146 22.832146 22.805004 22.590701 22.390829
#> 21 22.777513 22.777513 22.777513 22.777513 22.780947 22.808062 22.833352
#>       1.0580    0.9867    0.9202    0.8582   0.8003   0.7464   0.6961   0.6492
#> 1  21.984362 21.908444 21.837643 22.483997 23.66548 23.71730 23.70919 23.70697
#> 2  20.701612 20.784875 20.862525 20.570696 20.00272 20.00987 20.04536 20.05574
#> 3  19.557632 19.694770 19.822665 19.899925 19.93807 20.03542 20.12947 20.15721
#> 4  17.966044 17.828842 17.700887 16.738943 15.15696 14.92386 14.77335 14.72873
#> 5  21.840774 22.141168 22.421316 22.965639 23.70377 23.97455 24.20448 24.27239
#> 6  19.227981 19.243566 19.258100 18.832371 18.07843 18.02285 18.00582 18.00069
#> 7  19.364715 19.418166 19.468015 19.195123 18.68121 18.67268 18.69001 18.69504
#> 8  15.624136 15.594874 15.567585 15.350774 14.99304 14.94154 14.90870 14.89896
#> 9  17.503957 17.480866 17.459332 17.056992 16.37114 16.29503 16.25436 16.24227
#> 10 17.325371 17.328474 17.331367 17.065694 16.59988 16.56106 16.54613 16.54166
#> 11  8.415248  8.438183  8.459572  9.291474 10.72708 10.86898 10.93695 10.95720
#> 12  8.750284  8.746476  8.742924  9.417989 10.59875 10.69992 10.74050 10.75264
#> 13 26.184203 26.250100 26.311557 27.112522 28.46385 28.62779 28.72170 28.74960
#> 14 28.251309 28.337419 28.417725 29.074361 30.15957 30.31397 30.41181 30.44082
#> 15 24.877633 24.982638 25.080567 24.412297 23.17205 23.13502 23.16063 23.16800
#> 16 18.470512 18.397366 18.329149 17.594800 16.36484 16.20659 16.11222 16.08421
#> 17 16.411447 16.252230 16.103744 15.220055 13.79016 13.55534 13.39551 13.34817
#> 18 25.876638 25.766580 25.663938 25.717576 25.88877 25.82838 25.76028 25.74022
#> 19 29.304776 29.195443 29.093478 28.761974 28.26054 28.14152 28.04933 28.02207
#> 20 22.204427 22.030588 21.868466 22.054603 22.50199 22.42216 22.32106 22.29132
#> 21 22.856937 22.878932 22.899445 23.882196 25.58171 25.74617 25.82314 25.84609
#>      0.6054   0.5646   0.5266   0.4911   0.4580   0.4271   0.3983   0.3715
#> 1  23.70723 23.70738 23.70748 23.70753 23.70756 23.62764 23.49422 23.36975
#> 2  20.05560 20.05552 20.05547 20.05544 20.05542 20.16663 20.35273 20.52637
#> 3  20.15717 20.15715 20.15714 20.15713 20.15713 20.25343 20.41481 20.56540
#> 4  14.72846 14.72830 14.72820 14.72814 14.72811 14.69711 14.64449 14.59538
#> 5  24.27244 24.27247 24.27249 24.27250 24.27251 24.11150 23.84190 23.59034
#> 6  18.00053 18.00044 18.00039 18.00035 18.00034 17.84791 17.59214 17.35346
#> 7  18.69492 18.69485 18.69481 18.69478 18.69477 18.51535 18.21442 17.93361
#> 8  14.89890 14.89886 14.89884 14.89883 14.89882 14.95000 15.03563 15.11552
#> 9  16.24214 16.24206 16.24201 16.24199 16.24197 16.25612 16.27954 16.30138
#> 10 16.54156 16.54151 16.54148 16.54146 16.54144 16.54172 16.54197 16.54219
#> 11 10.95748 10.95764 10.95774 10.95780 10.95784 10.99060 11.04613 11.09795
#> 12 10.75288 10.75302 10.75310 10.75315 10.75318 10.79743 10.87210 10.94180
#> 13 28.74984 28.74999 28.75008 28.75013 28.75017 28.83781 28.98527 29.12288
#> 14 30.44101 30.44113 30.44119 30.44123 30.44126 30.47946 30.54394 30.60411
#> 15 23.16772 23.16755 23.16745 23.16740 23.16736 23.19108 23.23025 23.26680
#> 16 16.08399 16.08385 16.08378 16.08373 16.08370 16.22454 16.46006 16.67982
#> 17 13.34794 13.34780 13.34771 13.34766 13.34763 13.35824 13.37544 13.39147
#> 18 25.74029 25.74033 25.74036 25.74037 25.74038 25.83657 25.99788 26.14840
#> 19 28.02200 28.02196 28.02194 28.02193 28.02192 28.05741 28.11670 28.17203
#> 20 22.29146 22.29155 22.29160 22.29163 22.29165 22.04914 21.64298 21.26398
#> 21 25.84642 25.84662 25.84674 25.84681 25.84685 25.91031 26.01740 26.11734
#>      0.3464   0.3231   0.3013   0.2810   0.2621   0.2444   0.2279   0.2126
#> 1  23.25367 23.18817 23.18830 23.18838 23.15429 23.06315 23.00962 23.00928
#> 2  20.68831 20.77935 20.77938 20.77940 20.72317 20.57836 20.49345 20.49355
#> 3  20.70583 20.78465 20.78477 20.78483 20.63201 20.23506 20.00217 20.00202
#> 4  14.54958 14.52429 14.52399 14.52379 14.50894 14.47220 14.45043 14.45059
#> 5  23.35574 23.22410 23.22393 23.22383 23.24297 23.29263 23.32172 23.32168
#> 6  17.13087 17.00626 17.00590 17.00567 17.10122 17.35089 17.49723 17.49744
#> 7  17.67173 17.52503 17.52467 17.52444 17.60111 17.80138 17.91871 17.91885
#> 8  15.19004 15.23192 15.23193 15.23194 15.20696 15.14252 15.10471 15.10474
#> 9  16.32175 16.33337 16.33327 16.33320 16.27897 16.13889 16.05661 16.05663
#> 10 16.54241 16.54267 16.54258 16.54252 16.47908 16.31478 16.21828 16.21825
#> 11 11.14629 11.17299 11.17330 11.17350 11.16342 11.13501 11.11849 11.11826
#> 12 11.00679 11.04290 11.04318 11.04336 11.10421 11.26070 11.35272 11.35265
#> 13 29.25122 29.32284 29.32320 29.32343 29.34836 29.41149 29.44879 29.44868
#> 14 30.66023 30.69143 30.69166 30.69181 30.71733 30.78241 30.82078 30.82070
#> 15 23.30088 23.32041 23.32019 23.32004 23.36689 23.49134 23.56437 23.56472
#> 16 16.88476 17.00003 17.00002 17.00001 16.91268 16.68739 16.55518 16.55528
#> 17 13.40642 13.41514 13.41493 13.41479 13.57569 13.99600 14.24262 14.24306
#> 18 26.28878 26.36740 26.36760 26.36772 26.52871 26.94704 27.19274 27.19299
#> 19 28.22362 28.25265 28.25263 28.25262 28.18223 27.99945 27.89212 27.89205
#> 20 20.91053 20.71208 20.71186 20.71172 20.55293 20.13774 19.89369 19.89313
#> 21 26.21055 26.26234 26.26274 26.26300 26.31883 26.46155 26.54556 26.54542
#>      0.1982   0.1849   0.1724   0.1608   0.1500   0.1399   0.1304   0.1216
#> 1  23.00906 23.00890 23.00880 23.00873 23.00869 21.43502 21.43489 21.43476
#> 2  20.49362 20.49367 20.49370 20.49372 20.49374 20.59426 20.59429 20.59430
#> 3  20.00192 20.00186 20.00182 20.00180 20.00178 19.83975 19.83975 19.83975
#> 4  14.45071 14.45079 14.45085 14.45089 14.45092 14.36153 14.36156 14.36159
#> 5  23.32166 23.32164 23.32163 23.32163 23.32162 23.74843 23.74842 23.74842
#> 6  17.49759 17.49769 17.49776 17.49780 17.49783 18.11328 18.11324 18.11320
#> 7  17.91895 17.91902 17.91907 17.91911 17.91913 18.94315 18.94317 18.94319
#> 8  15.10477 15.10478 15.10479 15.10480 15.10481 15.37889 15.37896 15.37902
#> 9  16.05665 16.05667 16.05668 16.05669 16.05670 16.59030 16.59042 16.59052
#> 10 16.21824 16.21823 16.21823 16.21823 16.21823 17.00608 17.00623 17.00637
#> 11 11.11810 11.11799 11.11791 11.11786 11.11783 10.63331 10.63324 10.63317
#> 12 11.35259 11.35255 11.35252 11.35249 11.35247 11.39056 11.39054 11.39053
#> 13 29.44860 29.44854 29.44850 29.44847 29.44845 30.04229 30.04240 30.04250
#> 14 30.82064 30.82060 30.82056 30.82054 30.82052 32.07536 32.07557 32.07576
#> 15 23.56495 23.56511 23.56522 23.56529 23.56534 19.94405 19.94362 19.94321
#> 16 16.55535 16.55540 16.55544 16.55546 16.55548 16.63370 16.63375 16.63380
#> 17 14.24336 14.24356 14.24369 14.24377 14.24383 14.72405 14.72411 14.72418
#> 18 27.19315 27.19324 27.19330 27.19334 27.19336 26.61489 26.61476 26.61464
#> 19 27.89201 27.89198 27.89197 27.89196 27.89196 29.45043 29.45058 29.45072
#> 20 19.89277 19.89253 19.89238 19.89229 19.89223 19.03118 19.03106 19.03095
#> 21 26.54531 26.54523 26.54517 26.54513 26.54510 26.14949 26.14944 26.14941
#>      0.1134   0.1058   0.0987   0.0920   0.0858   0.0800   0.0746   0.0696
#> 1  21.43464 21.43452 21.43440 21.43429 21.43418 21.43408 21.43398 21.43388
#> 2  20.59432 20.59432 20.59433 20.59433 20.59434 20.59434 20.59433 20.59433
#> 3  19.83974 19.83974 19.83973 19.83972 19.83971 19.83970 19.83969 19.83968
#> 4  14.36163 14.36167 14.36171 14.36175 14.36179 14.36183 14.36187 14.36191
#> 5  23.74842 23.74842 23.74843 23.74845 23.74846 23.74848 23.74850 23.74852
#> 6  18.11315 18.11309 18.11304 18.11298 18.11292 18.11286 18.11281 18.11275
#> 7  18.94319 18.94319 18.94319 18.94318 18.94317 18.94315 18.94314 18.94312
#> 8  15.37907 15.37912 15.37916 15.37920 15.37923 15.37926 15.37929 15.37932
#> 9  16.59062 16.59071 16.59080 16.59088 16.59095 16.59102 16.59108 16.59113
#> 10 17.00651 17.00663 17.00675 17.00685 17.00695 17.00704 17.00713 17.00721
#> 11 10.63311 10.63306 10.63302 10.63298 10.63294 10.63291 10.63288 10.63285
#> 12 11.39053 11.39053 11.39053 11.39054 11.39055 11.39056 11.39057 11.39058
#> 13 30.04259 30.04268 30.04277 30.04285 30.04292 30.04299 30.04306 30.04312
#> 14 32.07595 32.07613 32.07630 32.07645 32.07660 32.07673 32.07686 32.07698
#> 15 19.94283 19.94248 19.94215 19.94184 19.94155 19.94127 19.94102 19.94078
#> 16 16.63383 16.63386 16.63388 16.63389 16.63391 16.63391 16.63392 16.63392
#> 17 14.72424 14.72430 14.72436 14.72442 14.72448 14.72454 14.72459 14.72464
#> 18 26.61453 26.61443 26.61433 26.61424 26.61416 26.61409 26.61402 26.61396
#> 19 29.45086 29.45099 29.45112 29.45124 29.45136 29.45147 29.45158 29.45168
#> 20 19.03085 19.03076 19.03068 19.03060 19.03052 19.03046 19.03039 19.03033
#> 21 26.14938 26.14936 26.14934 26.14933 26.14931 26.14931 26.14930 26.14929
#>      0.0649   0.0605   0.0565   0.0527   0.0491   0.0458   0.0427   0.0398
#> 1  21.43378 21.43369 21.43360 21.43352 21.43344 21.43336 21.43329 21.43322
#> 2  20.59433 20.59433 20.59433 20.59432 20.59432 20.59432 20.59432 20.59431
#> 3  19.83967 19.83966 19.83965 19.83965 19.83964 19.83963 19.83962 19.83962
#> 4  14.36194 14.36198 14.36201 14.36204 14.36206 14.36209 14.36211 14.36213
#> 5  23.74854 23.74856 23.74858 23.74860 23.74862 23.74864 23.74866 23.74868
#> 6  18.11270 18.11265 18.11260 18.11256 18.11251 18.11247 18.11243 18.11240
#> 7  18.94310 18.94309 18.94307 18.94305 18.94304 18.94302 18.94301 18.94299
#> 8  15.37934 15.37937 15.37939 15.37941 15.37942 15.37944 15.37945 15.37947
#> 9  16.59119 16.59123 16.59128 16.59132 16.59136 16.59139 16.59142 16.59145
#> 10 17.00728 17.00735 17.00741 17.00747 17.00752 17.00757 17.00762 17.00766
#> 11 10.63283 10.63281 10.63279 10.63278 10.63276 10.63275 10.63274 10.63273
#> 12 11.39060 11.39061 11.39063 11.39064 11.39066 11.39067 11.39069 11.39070
#> 13 30.04317 30.04322 30.04327 30.04332 30.04336 30.04340 30.04343 30.04346
#> 14 32.07709 32.07719 32.07728 32.07737 32.07745 32.07753 32.07759 32.07766
#> 15 19.94056 19.94036 19.94016 19.93998 19.93982 19.93966 19.93951 19.93938
#> 16 16.63393 16.63393 16.63393 16.63393 16.63393 16.63393 16.63393 16.63393
#> 17 14.72469 14.72473 14.72477 14.72481 14.72485 14.72488 14.72492 14.72495
#> 18 26.61391 26.61386 26.61382 26.61378 26.61375 26.61372 26.61370 26.61368
#> 19 29.45178 29.45187 29.45196 29.45204 29.45212 29.45220 29.45227 29.45234
#> 20 19.03028 19.03022 19.03017 19.03013 19.03008 19.03004 19.03000 19.02996
#> 21 26.14929 26.14929 26.14928 26.14928 26.14928 26.14928 26.14928 26.14928
#>      0.0371   0.0346   0.0323   0.0301   0.0281   0.0262   0.0244   0.0228
#> 1  21.68634 21.68656 21.68676 21.68695 21.68713 21.68730 21.68746 21.68761
#> 2  20.92025 20.92028 20.92031 20.92033 20.92035 20.92036 20.92038 20.92038
#> 3  19.73788 19.73787 19.73786 19.73786 19.73785 19.73785 19.73784 19.73784
#> 4  14.63717 14.63736 14.63754 14.63770 14.63785 14.63799 14.63812 14.63824
#> 5  23.85607 23.85612 23.85616 23.85620 23.85624 23.85627 23.85630 23.85633
#> 6  18.05691 18.05676 18.05663 18.05651 18.05639 18.05629 18.05619 18.05610
#> 7  18.95351 18.95336 18.95323 18.95310 18.95298 18.95286 18.95276 18.95266
#> 8  14.89032 14.89013 14.88996 14.88979 14.88965 14.88951 14.88938 14.88926
#> 9  16.38046 16.38034 16.38023 16.38013 16.38003 16.37994 16.37986 16.37978
#> 10 16.80301 16.80288 16.80276 16.80264 16.80253 16.80243 16.80233 16.80224
#> 11 10.89412 10.89424 10.89435 10.89445 10.89455 10.89464 10.89472 10.89480
#> 12 11.39261 11.39261 11.39261 11.39261 11.39261 11.39261 11.39261 11.39262
#> 13 29.95419 29.95408 29.95398 29.95388 29.95378 29.95370 29.95361 29.95353
#> 14 32.27140 32.27132 32.27123 32.27115 32.27107 32.27099 32.27091 32.27083
#> 15 19.80896 19.80937 19.80977 19.81015 19.81052 19.81086 19.81120 19.81152
#> 16 16.82279 16.82271 16.82264 16.82257 16.82250 16.82244 16.82237 16.82231
#> 17 14.74589 14.74586 14.74584 14.74582 14.74580 14.74578 14.74576 14.74574
#> 18 26.62717 26.62708 26.62699 26.62690 26.62683 26.62676 26.62669 26.62663
#> 19 29.49027 29.49014 29.49003 29.48991 29.48980 29.48970 29.48959 29.48949
#> 20 18.91892 18.91912 18.91931 18.91950 18.91967 18.91984 18.92000 18.92015
#> 21 25.85180 25.85182 25.85183 25.85185 25.85187 25.85189 25.85191 25.85193
#>      0.0213   0.0198   0.0185   0.0172   0.0161   0.0150   0.0140   0.0130
#> 1  21.68775 21.68789 21.68802 21.68813 21.68825 21.68835 21.68845 21.68854
#> 2  20.92039 20.92040 20.92040 20.92040 20.92040 20.92039 20.92039 20.92038
#> 3  19.73784 19.73784 19.73784 19.73784 19.73784 19.73784 19.73784 19.73784
#> 4  14.63835 14.63845 14.63855 14.63863 14.63871 14.63877 14.63884 14.63889
#> 5  23.85635 23.85637 23.85639 23.85641 23.85643 23.85644 23.85646 23.85647
#> 6  18.05602 18.05594 18.05587 18.05581 18.05575 18.05570 18.05565 18.05560
#> 7  18.95256 18.95248 18.95239 18.95232 18.95225 18.95218 18.95212 18.95207
#> 8  14.88916 14.88906 14.88897 14.88888 14.88881 14.88874 14.88868 14.88862
#> 9  16.37970 16.37963 16.37957 16.37951 16.37945 16.37939 16.37934 16.37929
#> 10 16.80215 16.80207 16.80199 16.80192 16.80185 16.80178 16.80172 16.80166
#> 11 10.89488 10.89494 10.89500 10.89506 10.89511 10.89516 10.89521 10.89525
#> 12 11.39262 11.39262 11.39262 11.39262 11.39262 11.39262 11.39262 11.39262
#> 13 29.95346 29.95338 29.95332 29.95325 29.95319 29.95314 29.95308 29.95303
#> 14 32.27075 32.27067 32.27060 32.27053 32.27046 32.27039 32.27032 32.27026
#> 15 19.81182 19.81212 19.81240 19.81266 19.81291 19.81316 19.81339 19.81361
#> 16 16.82225 16.82220 16.82214 16.82209 16.82204 16.82199 16.82195 16.82190
#> 17 14.74572 14.74570 14.74568 14.74566 14.74564 14.74563 14.74561 14.74559
#> 18 26.62658 26.62653 26.62649 26.62644 26.62641 26.62637 26.62634 26.62632
#> 19 29.48940 29.48931 29.48922 29.48914 29.48905 29.48898 29.48890 29.48883
#> 20 18.92029 18.92043 18.92056 18.92068 18.92080 18.92091 18.92101 18.92111
#> 21 25.85195 25.85197 25.85200 25.85202 25.85204 25.85207 25.85209 25.85212
#>      0.0122   0.0113   0.0106   0.0099   0.0092   0.0086   0.0080   0.0075
#> 1  21.68863 21.68871 21.68879 21.68886 21.68893 21.68899 21.68905 21.68910
#> 2  20.92038 20.92037 20.92036 20.92035 20.92034 20.92033 20.92032 20.92030
#> 3  19.73784 19.73785 19.73785 19.73785 19.73785 19.73786 19.73786 19.73786
#> 4  14.63894 14.63899 14.63903 14.63907 14.63910 14.63913 14.63915 14.63917
#> 5  23.85648 23.85649 23.85649 23.85650 23.85650 23.85651 23.85651 23.85651
#> 6  18.05556 18.05553 18.05549 18.05546 18.05544 18.05542 18.05540 18.05538
#> 7  18.95201 18.95196 18.95192 18.95188 18.95184 18.95180 18.95177 18.95174
#> 8  14.88857 14.88852 14.88848 14.88844 14.88841 14.88838 14.88836 14.88833
#> 9  16.37925 16.37921 16.37917 16.37913 16.37910 16.37907 16.37904 16.37901
#> 10 16.80161 16.80156 16.80151 16.80147 16.80142 16.80138 16.80135 16.80131
#> 11 10.89529 10.89532 10.89535 10.89538 10.89541 10.89543 10.89545 10.89547
#> 12 11.39262 11.39262 11.39262 11.39262 11.39261 11.39261 11.39261 11.39261
#> 13 29.95298 29.95294 29.95290 29.95286 29.95282 29.95279 29.95275 29.95272
#> 14 32.27019 32.27013 32.27007 32.27001 32.26995 32.26990 32.26985 32.26980
#> 15 19.81382 19.81401 19.81420 19.81438 19.81455 19.81472 19.81487 19.81502
#> 16 16.82186 16.82182 16.82178 16.82175 16.82171 16.82168 16.82164 16.82161
#> 17 14.74557 14.74556 14.74554 14.74553 14.74551 14.74550 14.74548 14.74547
#> 18 26.62629 26.62627 26.62625 26.62623 26.62622 26.62621 26.62620 26.62619
#> 19 29.48876 29.48869 29.48863 29.48857 29.48851 29.48846 29.48840 29.48835
#> 20 18.92120 18.92129 18.92138 18.92145 18.92153 18.92160 18.92167 18.92173
#> 21 25.85214 25.85216 25.85219 25.85221 25.85223 25.85226 25.85228 25.85230
#>      0.0070   0.0065   0.0061   0.0056   0.0053
#> 1  21.68915 21.68920 21.68924 21.68928 21.68932
#> 2  20.92029 20.92028 20.92027 20.92025 20.92024
#> 3  19.73786 19.73787 19.73787 19.73787 19.73788
#> 4  14.63919 14.63920 14.63922 14.63923 14.63923
#> 5  23.85651 23.85651 23.85651 23.85651 23.85651
#> 6  18.05536 18.05535 18.05534 18.05533 18.05532
#> 7  18.95171 18.95169 18.95167 18.95165 18.95163
#> 8  14.88832 14.88830 14.88829 14.88828 14.88827
#> 9  16.37899 16.37897 16.37895 16.37893 16.37891
#> 10 16.80128 16.80125 16.80122 16.80120 16.80117
#> 11 10.89549 10.89550 10.89551 10.89553 10.89554
#> 12 11.39261 11.39261 11.39261 11.39261 11.39261
#> 13 29.95269 29.95267 29.95264 29.95262 29.95260
#> 14 32.26975 32.26970 32.26965 32.26961 32.26956
#> 15 19.81516 19.81529 19.81541 19.81553 19.81564
#> 16 16.82158 16.82156 16.82153 16.82151 16.82148
#> 17 14.74545 14.74544 14.74543 14.74542 14.74540
#> 18 26.62618 26.62617 26.62617 26.62617 26.62616
#> 19 29.48830 29.48826 29.48821 29.48817 29.48813
#> 20 18.92179 18.92184 18.92189 18.92194 18.92199
#> 21 25.85232 25.85234 25.85237 25.85239 25.85241
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.4 18.1 14.3 22.8 19.2 17.8 16.4 17.3 15.2 10.4 14.7 32.4 33.9 21.5
#> [16] 15.5 13.3 26.0 30.4 19.7 21.4
#> 
#> $X
#>               am       carb         cyl        disp       drat       gear
#>  [1,]  1.4142136  0.8992288 -0.05643326 -0.53021082  0.7055269  0.5275044
#>  [2,] -0.7071068 -1.2796718 -0.05643326  0.33190016 -0.9519319 -0.8571946
#>  [3,] -0.7071068 -1.2796718 -0.05643326  0.04159748 -1.5987451 -0.8571946
#>  [4,] -0.7071068  0.8992288  1.12866530  1.22919934 -0.6891641 -0.8571946
#>  [5,] -0.7071068 -0.5533716 -1.24153183 -0.69911419  0.7459527  0.5275044
#>  [6,] -0.7071068  0.8992288 -0.05643326 -0.46335323  0.7459527  0.5275044
#>  [7,] -0.7071068  0.8992288 -0.05643326 -0.46335323  0.7459527  0.5275044
#>  [8,] -0.7071068  0.1729286  1.12866530  0.48848766 -0.9721448 -0.8571946
#>  [9,] -0.7071068  0.1729286  1.12866530  0.48848766 -0.9721448 -0.8571946
#> [10,] -0.7071068  0.1729286  1.12866530  0.48848766 -0.9721448 -0.8571946
#> [11,] -0.7071068  0.8992288  1.12866530  2.10890441 -1.1136352 -0.8571946
#> [12,] -0.7071068  0.8992288  1.12866530  1.93296340 -0.6487382 -0.8571946
#> [13,]  1.4142136 -1.2796718 -1.24153183 -1.24541104  1.0693593  0.5275044
#> [14,]  1.4142136 -1.2796718 -1.24153183 -1.31226863  1.3523401  0.5275044
#> [15,] -0.7071068 -1.2796718 -1.24153183 -0.88121314  0.3012686 -0.8571946
#> [16,] -0.7071068 -0.5533716  1.12866530  0.85972320 -1.5987451 -0.8571946
#> [17,] -0.7071068  0.8992288  1.12866530  1.14122883  0.3619074 -0.8571946
#> [18,]  1.4142136 -0.5533716 -1.24153183 -0.87945373  1.7768112  1.9122034
#> [19,]  1.4142136 -0.5533716 -1.24153183 -1.10113941  0.4427590  1.9122034
#> [20,]  1.4142136  2.3518292 -0.05643326 -0.66216658  0.1395653  1.9122034
#> [21,]  1.4142136 -0.5533716 -1.24153183 -0.87329580  1.1299981  0.5275044
#>               hp        qsec         vs          wt
#>  [1,] -0.5995580 -0.61476736 -0.9534626 -0.40449678
#>  [2,] -0.5995580  0.75551673  1.0488088 -0.05172934
#>  [3,] -0.6898788  1.19717854  1.0488088  0.20247072
#>  [4,]  1.8391032 -1.28292241 -0.9534626  0.31660136
#>  [5,] -0.8705203  2.71468324  1.0488088 -0.11917017
#>  [6,] -0.3647239  0.11001100  1.0488088  0.18171969
#>  [7,] -0.3647239  0.44975086  1.0488088  0.18171969
#>  [8,]  0.6649330 -0.39959878 -0.9534626  0.83537700
#>  [9,]  0.6649330 -0.28635217 -0.9534626  0.48260956
#> [10,]  0.6649330 -0.05985893 -0.9534626  0.53448713
#> [11,]  1.2971785 -0.16178088 -0.9534626  2.24022142
#> [12,]  1.5681409 -0.38827412 -0.9534626  2.15825487
#> [13,] -1.3943809  0.77250372  1.0488088 -1.10484388
#> [14,] -1.4124450  1.01598395  1.0488088 -1.48355010
#> [15,] -0.8343920  1.07826959  1.0488088 -0.82989280
#> [16,]  0.1230083 -0.69970232 -0.9534626  0.26472380
#> [17,]  1.8391032 -1.52640264 -0.9534626  0.59674020
#> [18,] -0.9427770 -0.79596195 -0.9534626 -1.16709696
#> [19,] -0.5453655 -0.68271533  1.0488088 -1.81764161
#> [20,]  0.5746122 -1.47544166 -0.9534626 -0.51343966
#> [21,] -0.6176221  0.27988093  1.0488088 -0.50306415
#> attr(,"center")
#>  [1]   0.3333333   2.7619048   6.0952381 220.2714286   3.5509524   3.6190476
#>  [7] 143.1904762  18.1057143   0.4761905   3.2648571
#> attr(,"scale")
#>  [1]   0.4714045   1.3768412   1.6876233 113.6744606   0.4947333   0.7221786
#>  [7]  55.3582437   1.7660571   0.4994328   0.9638078
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 32
#> 
#> $lambda.min
#> [1] 0.6054175
#> 
#> $null.dev
#> [1] 36.95918
#> 
#> $Bias
#> [1] 7.373522
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 9.100254