Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 42.500961 40.289959 34.892141 30.355441 26.557046 23.390856 20.765263
#>   [8] 18.601212 16.830529 15.394464 14.242424 13.330878 12.622407 12.084875
#>  [15] 11.690716 11.416311 11.209918 11.036988 11.049785 10.884462 10.556892
#>  [22] 10.304430 10.115103  9.751094  7.744161  7.205867  7.043781  6.919439
#>  [29]  6.826794  6.760670  6.727244  6.725224  6.736949  6.731853  6.732315
#>  [36]  6.710123  6.675181  6.695515  6.982787  7.077527  7.388728  7.475313
#>  [43]  7.579424  7.598875  7.804554  7.751597  8.792652  8.727467  8.770346
#>  [50] 11.174896 11.571519 11.578488 11.547557 12.453293 13.268849 13.191234
#>  [57] 14.269495 14.354259 14.331080 15.367548 15.390603 15.447209 15.338351
#>  [64] 15.339160 16.167568 16.102514 15.840538 15.841315 15.901688 15.902488
#>  [71] 15.904757 15.351063 15.367023 15.369039 15.369608 15.367231 15.394059
#>  [78] 15.394584 15.395094 15.395590 15.396073 15.396542 15.391005 15.225660
#>  [85] 15.443393 15.444380 14.977825 14.975028 14.930800 14.929763 14.807097
#>  [92] 14.805681 14.804338 14.803082 14.800912 14.799756 14.798661 14.797626
#>  [99] 14.796648 14.795722
#> 
#> $cvse
#>   [1] 11.133067 10.864164  9.558792  8.464358  7.543502  6.764650  6.101600
#>   [8]  5.533069  5.042202  4.616016  4.244837  3.921722  3.641909  3.402281
#>  [15]  3.200862  3.036317  2.908076  2.825330  2.824417  2.834944  2.860630
#>  [22]  2.884579  2.904331  2.896235  2.730364  2.634821  2.564205  2.499167
#>  [29]  2.439763  2.386079  2.340169  2.302745  2.271071  2.243667  2.220773
#>  [36]  2.196969  2.173733  2.159455  2.211683  2.201633  2.196683  2.200113
#>  [43]  2.174659  2.169916  2.157281  2.127985  2.568563  2.536107  2.492489
#>  [50]  3.564576  3.671083  3.670665  3.685051  4.058060  4.088294  4.093807
#>  [57]  4.083912  4.013798  4.017083  4.542686  4.539699  4.518967  4.531240
#>  [64]  4.531583  5.242761  5.232506  5.231004  5.231412  5.253385  5.253786
#>  [71]  5.254098  5.254610  5.253061  5.253251  5.253611  5.254555  5.257706
#>  [78]  5.258057  5.258400  5.258733  5.259057  5.259373  5.260385  5.246237
#>  [85]  5.250067  5.250363  4.903885  4.902951  4.896118  4.895011  4.887579
#>  [92]  4.886559  4.885587  4.884662  4.883962  4.883124  4.882326  4.881566
#>  [99]  4.880843  4.880154
#> 
#> $fold
#>  [1]  8  9  6  4  3  4  5  2  9  1 10  2  7 10  5  1  8  6  1  7  3
#> 
#> $lambda
#>   [1] 5.649119690 5.268387930 4.913316216 4.582175147 4.273351878 3.985342264
#>   [7] 3.716743534 3.466247459 3.232633982 3.014765271 2.811580181 2.622089087
#>  [13] 2.445369059 2.280559368 2.126857300 1.983514236 1.849832015 1.725159528
#>  [19] 1.608889550 1.500455779 1.399330082 1.305019917 1.217065943 1.135039772
#>  [25] 1.058541890 0.987199709 0.920665753 0.858615963 0.800748120 0.746780377
#>  [31] 0.696449879 0.649511488 0.605736588 0.564911969 0.526838793 0.491331622
#>  [37] 0.458217515 0.427335188 0.398534226 0.371674353 0.346624746 0.323263398
#>  [43] 0.301476527 0.281158018 0.262208909 0.244536906 0.228055937 0.212685730
#>  [49] 0.198351424 0.184983202 0.172515953 0.160888955 0.150045578 0.139933008
#>  [55] 0.130501992 0.121706594 0.113503977 0.105854189 0.098719971 0.092066575
#>  [61] 0.085861596 0.080074812 0.074678038 0.069644988 0.064951149 0.060573659
#>  [67] 0.056491197 0.052683879 0.049133162 0.045821751 0.042733519 0.039853423
#>  [73] 0.037167435 0.034662475 0.032326340 0.030147653 0.028115802 0.026220891
#>  [79] 0.024453691 0.022805594 0.021268573 0.019835142 0.018498320 0.017251595
#>  [85] 0.016088896 0.015004558 0.013993301 0.013050199 0.012170659 0.011350398
#>  [91] 0.010585419 0.009871997 0.009206658 0.008586160 0.008007481 0.007467804
#>  [97] 0.006964499 0.006495115 0.006057366 0.005649120
#> 
#> $fit
#> $beta
#>               5.6491     5.2684    4.9133    4.5822    4.2734    3.9853
#> (Intercept) 19.94286 21.7435665 23.422914 24.989079 26.449690 27.811861
#> am           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000 -0.5691843 -1.100007 -1.595055 -2.056738 -2.487305
#>                3.7167    3.4662    3.2326    3.0148    2.8116    2.6221
#> (Intercept) 29.082226 30.266972 31.371871 32.402303 33.363287 34.259505
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.888853 -3.263338 -3.612585 -3.938293 -4.242049 -4.525334
#>                2.4454    2.2806    2.1269    1.9835    1.8498    1.7252
#> (Intercept) 35.095320 35.874804 36.601753 37.279709 37.754913 37.754913
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -4.789526 -5.035912 -5.265693 -5.479987 -5.630193 -5.630193
#>                1.6089      1.5005     1.3993     1.3050     1.2171     1.1350
#> (Intercept) 37.754913 36.26027025 34.7898213 33.4184757 32.1395541 30.9468276
#> am           0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.000000  0.08271875  0.1640986  0.2399936  0.3107737  0.3767833
#> vs           0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.630193 -5.61722198 -5.6044605 -5.5925590 -5.5814597 -5.5711084
#>                 1.0585    0.9872     0.9207     0.8586     0.8007     0.7468
#> (Intercept) 29.8344869 28.797114 27.8296569 26.9274031 26.0859581 25.3012237
#> am           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.4383442  0.495756  0.5492985  0.5992324  0.6458009  0.6892309
#> vs           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.5614548 -5.552452 -5.5440556 -5.5362252 -5.5289226 -5.5221121
#>                 0.6964    0.6495     0.6057     0.5649     0.5268     0.4913
#> (Intercept) 24.5693778 23.886856 23.2503335 22.6567107 22.2678796 22.2678608
#> am           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.7297338  0.767507  0.8027343  0.8355875  0.8571068  0.8571078
#> vs           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.5157607 -5.509837 -5.5043132 -5.4991613 -5.4957868 -5.4957866
#>                 0.4582     0.4273     0.3985     0.3717     0.3466     0.3233
#> (Intercept) 22.2678606 22.2678606 22.2678606 22.2678606 22.2678606 22.2678606
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.8571078  0.8571078  0.8571078  0.8571078  0.8571078  0.8571078
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.4957866 -5.4957866 -5.4957866 -5.4957866 -5.4957866 -5.4957866
#>                 0.3015    0.2812     0.2622    0.2445     0.2281     0.2127
#> (Intercept) 40.6588469 40.659706 40.6599279 40.660075 40.6601490 40.6601475
#> am           0.0000000  0.000000  0.0000000  0.000000  0.0000000  0.0000000
#> carb        -0.3200712 -0.345787 -0.3701495 -0.392899 -0.4048846 -0.4049073
#> cyl         -1.4437422 -1.432251 -1.4210402 -1.410539 -1.4050039 -1.4049925
#> disp         0.0000000  0.000000  0.0000000  0.000000  0.0000000  0.0000000
#> drat         0.0000000  0.000000  0.0000000  0.000000  0.0000000  0.0000000
#> gear         0.0000000  0.000000  0.0000000  0.000000  0.0000000  0.0000000
#> hp           0.0000000  0.000000  0.0000000  0.000000  0.0000000  0.0000000
#> qsec         0.0000000  0.000000  0.0000000  0.000000  0.0000000  0.0000000
#> vs           0.0000000  0.000000  0.0000000  0.000000  0.0000000  0.0000000
#> wt          -3.3857260 -3.385217 -3.3851928 -3.385216 -3.3852320 -3.3852333
#>                 0.1984     0.1850    0.1725     0.1609     0.1500     0.1399
#> (Intercept) 40.6601465 40.6601458 40.660145 40.6601451 40.6601449 40.6601448
#> am           0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> carb        -0.4049124 -0.4049137 -0.404914 -0.4049141 -0.4049142 -0.4049142
#> cyl         -1.4049895 -1.4049886 -1.404988 -1.4049880 -1.4049879 -1.4049878
#> disp         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> vs           0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> wt          -3.3852342 -3.3852347 -3.385235 -3.3852352 -3.3852354 -3.3852355
#>                 0.1305     0.1217     0.1135      0.1059      0.0987
#> (Intercept) 40.6601447 40.6601447 35.0499941 35.04903443 35.04812031
#> am           0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> carb        -0.4049142 -0.4049142  0.6729025  0.67301563  0.67312344
#> cyl         -1.4049877 -1.4049877 -1.4965771 -1.49667206 -1.49676300
#> disp         0.0000000  0.0000000  0.0328020  0.03280658  0.03281095
#> drat         0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> gear         0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> hp           0.0000000  0.0000000 -0.0246315 -0.02463248 -0.02463340
#> qsec         0.0000000  0.0000000  0.4350986  0.43519283  0.43528274
#> vs           0.0000000  0.0000000  1.1362763  1.13621297  1.13615200
#> wt          -3.3852355 -3.3852355 -6.2553225 -6.25574776 -6.25615321
#>                  0.0921      0.0859      0.0801      0.0747      0.0696
#> (Intercept) 35.04724966 35.04642046 35.04563079 35.04487879 35.04416271
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.67322616  0.67332401  0.67341723  0.67350602  0.67359059
#> cyl         -1.49685009 -1.49693347 -1.49701328 -1.49708965 -1.49716270
#> disp         0.03281512  0.03281910  0.03282290  0.03282651  0.03282996
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.02463428 -0.02463511 -0.02463589 -0.02463664 -0.02463735
#> qsec         0.43536850  0.43545029  0.43552828  0.43560265  0.43567355
#> vs           1.13609332  1.13603689  1.13598265  1.13593054  1.13588050
#> wt          -6.25653972 -6.25690812 -6.25725924 -6.25759387 -6.25791274
#>                  0.0650      0.0606      0.0565      0.0527      0.0491
#> (Intercept) 31.38560462 31.38296657 31.38056017 31.37836403 31.37635881
#> am           0.82877907  0.82941218  0.82999560  0.83053362  0.83103014
#> carb         0.63039222  0.63027697  0.63016854  0.63006648  0.62997035
#> cyl         -1.27187030 -1.27154616 -1.27124609 -1.27096810 -1.27071039
#> disp         0.03215287  0.03214880  0.03214496  0.03214133  0.03213789
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.02580152 -0.02580165 -0.02580176 -0.02580185 -0.02580191
#> qsec         0.53351543  0.53353461  0.53355019  0.53356261  0.53357224
#> vs           1.20382618  1.20402730  1.20421332  1.20438550  1.20454499
#> wt          -6.07924031 -6.07885871 -6.07849771 -6.07815609 -6.07783272
#>                  0.0458      0.0427      0.0399      0.0372      0.0347
#> (Intercept) 31.37452700 31.37285277 31.37132177 31.36992101 31.36863874
#> am           0.83148868  0.83191247  0.83230443  0.83266720  0.83300320
#> carb         0.62987977  0.62979439  0.62971386  0.62963788  0.62956615
#> cyl         -1.27047130 -1.27024934 -1.27004312 -1.26985139 -1.26967302
#> disp         0.03213465  0.03213158  0.03212868  0.03212594  0.03212334
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.02580196 -0.02580199 -0.02580202 -0.02580203 -0.02580203
#> qsec         0.53357943  0.53358450  0.53358771  0.53358931  0.53358952
#> vs           1.20469285  1.20483002  1.20495738  1.20507570  1.20518572
#> wt          -6.07752651 -6.07723649 -6.07696171 -6.07670131 -6.07645448
#>                  0.0323      0.0301      0.0281      0.0262      0.0245
#> (Intercept) 33.36891664 33.36993427 33.37082740 33.37160331 33.37226898
#> am           0.90966858  0.90959103  0.90952217  0.90946150  0.90940853
#> carb         0.68429871  0.68415160  0.68400755  0.68386660  0.68372878
#> cyl         -1.39960651 -1.39958251 -1.39955073 -1.39951179 -1.39946625
#> disp         0.03247492  0.03247022  0.03246563  0.03246116  0.03245681
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear        -0.32181543 -0.32174165 -0.32166168 -0.32157611 -0.32148550
#> hp          -0.02397564 -0.02397370 -0.02397193 -0.02397030 -0.02396882
#> qsec         0.52681833  0.52672241  0.52663015  0.52654146  0.52645624
#> vs           1.14635324  1.14628530  1.14622639  1.14617596  1.14613349
#> wt          -6.19404029 -6.19355004 -6.19307031 -6.19260120 -6.19214280
#>                  0.0228      0.0213      0.0198        0.0185      0.0173
#> (Intercept) 33.37283109 33.37329604 33.37366994  3.337392e+01 32.83990584
#> am           0.90936283  0.90932395  0.90929147  9.092650e-01  0.90129450
#> carb         0.68359411  0.68346262  0.68333430  6.832092e-01  0.66507623
#> cyl         -1.39941468 -1.39935760 -1.39929551 -1.399229e+00 -1.37431775
#> disp         0.03245258  0.03244846  0.03244447  3.244058e-02  0.03200565
#> drat         0.00000000  0.00000000  0.00000000  1.448802e-05  0.10192251
#> gear        -0.32139037 -0.32129122 -0.32118851 -3.210897e-01 -0.32220693
#> hp          -0.02396747 -0.02396624 -0.02396514 -2.396411e-02 -0.02378470
#> qsec         0.52637439  0.52629583  0.52622046  5.261484e-01  0.52453458
#> vs           1.14609850  1.14607051  1.14604906  1.146033e+00  1.14152665
#> wt          -6.19169515 -6.19125829 -6.19083224 -6.190417e+00 -6.13477010
#>                  0.0161      0.0150      0.0140      0.0131      0.0122
#> (Intercept) 32.83616639 32.83253030 32.82899667 32.82556444 32.82223244
#> am           0.90147604  0.90165402  0.90182835  0.90199895  0.90216578
#> carb         0.66496490  0.66485918  0.66475887  0.66466377  0.66457369
#> cyl         -1.37397875 -1.37365032 -1.37333229 -1.37302448 -1.37272669
#> disp         0.03200306  0.03200062  0.03199833  0.03199617  0.03199414
#> drat         0.10205398  0.10217948  0.10229919  0.10241331  0.10252202
#> gear        -0.32188480 -0.32157305 -0.32127146 -0.32097983 -0.32069796
#> hp          -0.02378751 -0.02379028 -0.02379299 -0.02379564 -0.02379824
#> qsec         0.52453376  0.52453441  0.52453642  0.52453970  0.52454414
#> vs           1.14174808  1.14196417  1.14217491  1.14238034  1.14258048
#> wt          -6.13441774 -6.13408359 -6.13376695 -6.13346716 -6.13318359
#>                  0.0114      0.0106      0.0099      0.0092      0.0086
#> (Intercept) 32.81899941 32.81586397 32.81282469 32.80988006 32.80702849
#> am           0.90232879  0.90248796  0.90264328  0.90279474  0.90294235
#> carb         0.66448845  0.66440785  0.66433172  0.66425989  0.66419219
#> cyl         -1.37243875 -1.37216046 -1.37189161 -1.37163202 -1.37138148
#> disp         0.03199224  0.03199046  0.03198880  0.03198725  0.03198580
#> drat         0.10262551  0.10272396  0.10281755  0.10290646  0.10299085
#> gear        -0.32042564 -0.32016266 -0.31990882 -0.31966391 -0.31942772
#> hp          -0.02380079 -0.02380328 -0.02380572 -0.02380811 -0.02381043
#> qsec         0.52454967  0.52455619  0.52456363  0.52457191  0.52458097
#> vs           1.14277536  1.14296503  1.14314953  1.14332892  1.14350325
#> wt          -6.13291561 -6.13266260 -6.13242397 -6.13219916 -6.13198760
#>                  0.0080      0.0075      0.0070      0.0065      0.0061
#> (Intercept) 32.80426835 32.80159799 32.79901567 32.79651966 32.79410819
#> am           0.90308611  0.90322606  0.90336220  0.90349458  0.90362323
#> carb         0.66412844  0.66406850  0.66401220  0.66395940  0.66390995
#> cyl         -1.37113978 -1.37090672 -1.37068210 -1.37046572 -1.37025736
#> disp         0.03198445  0.03198320  0.03198205  0.03198098  0.03197999
#> drat         0.10307088  0.10314673  0.10321855  0.10328648  0.10335069
#> gear        -0.31920004 -0.31898067 -0.31876939 -0.31856601 -0.31837032
#> hp          -0.02381271 -0.02381492 -0.02381709 -0.02381919 -0.02382125
#> qsec         0.52459072  0.52460112  0.52461210  0.52462361  0.52463558
#> vs           1.14367259  1.14383700  1.14399656  1.14415132  1.14430138
#> wt          -6.13178874 -6.13160207 -6.13142707 -6.13126325 -6.13111013
#>                  0.0056
#> (Intercept) 32.79177947
#> am           0.90374819
#> carb         0.66386370
#> cyl         -1.37005683
#> disp         0.03197909
#> drat         0.10341131
#> gear        -0.31818212
#> hp          -0.02382325
#> qsec         0.52464797
#> vs           1.14444681
#> wt          -6.13096723
#> 
#> $iter
#>   [1]   0   5   3   3   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   4   3   3   3   3   3   3   3   3   3   3   3   3   3   3   2   1
#>  [37]   1   1   1   1   1   1 109   5   5   5   4   1   1   1   1   1   1   1
#>  [55]   1   1 310   1   1   1   1   1   1   1  54   1   1   1   1   1   1   1
#>  [73]   1   1  91   1   1   1   1   1   1   1   1  63   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.649119690 5.268387930 4.913316216 4.582175147 4.273351878 3.985342264
#>   [7] 3.716743534 3.466247459 3.232633982 3.014765271 2.811580181 2.622089087
#>  [13] 2.445369059 2.280559368 2.126857300 1.983514236 1.849832015 1.725159528
#>  [19] 1.608889550 1.500455779 1.399330082 1.305019917 1.217065943 1.135039772
#>  [25] 1.058541890 0.987199709 0.920665753 0.858615963 0.800748120 0.746780377
#>  [31] 0.696449879 0.649511488 0.605736588 0.564911969 0.526838793 0.491331622
#>  [37] 0.458217515 0.427335188 0.398534226 0.371674353 0.346624746 0.323263398
#>  [43] 0.301476527 0.281158018 0.262208909 0.244536906 0.228055937 0.212685730
#>  [49] 0.198351424 0.184983202 0.172515953 0.160888955 0.150045578 0.139933008
#>  [55] 0.130501992 0.121706594 0.113503977 0.105854189 0.098719971 0.092066575
#>  [61] 0.085861596 0.080074812 0.074678038 0.069644988 0.064951149 0.060573659
#>  [67] 0.056491197 0.052683879 0.049133162 0.045821751 0.042733519 0.039853423
#>  [73] 0.037167435 0.034662475 0.032326340 0.030147653 0.028115802 0.026220891
#>  [79] 0.024453691 0.022805594 0.021268573 0.019835142 0.018498320 0.017251595
#>  [85] 0.016088896 0.015004558 0.013993301 0.013050199 0.012170659 0.011350398
#>  [91] 0.010585419 0.009871997 0.009206658 0.008586160 0.008007481 0.007467804
#>  [97] 0.006964499 0.006495115 0.006057366 0.005649120
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 42
#> 
#> $loss
#>   [1] 820.83143 692.18028 584.54446 494.89970 420.63500 359.49747 309.54447
#>   [8] 269.10212 236.72916 211.18572 191.40615 176.47546 165.60893 158.13434
#>  [15] 153.47668 151.14480 150.66781 150.66781 150.66781 140.68209 131.84613
#>  [22] 124.48889 118.39567 113.38122 109.28583 105.97185 103.32080 101.23073
#>  [29]  99.61406  98.39555  97.51071  96.90427  96.52899  96.34450  96.31024
#>  [36]  96.31024  96.31024  96.31024  96.31024  96.31024  96.31024  96.31024
#>  [43]  81.48152  81.29747  81.18376  81.13086  81.12369  81.12369  81.12369
#>  [50]  81.12369  81.12369  81.12369  81.12369  81.12369  81.12369  81.12369
#>  [57]  67.28310  67.28309  67.28308  67.28307  67.28306  67.28305  67.28304
#>  [64]  67.28303  66.68379  66.68377  66.68375  66.68374  66.68373  66.68371
#>  [71]  66.68370  66.68370  66.68369  66.68368  66.51114  66.51112  66.51111
#>  [78]  66.51110  66.51108  66.51107  66.51106  66.51105  66.51103  66.48712
#>  [85]  66.48711  66.48709  66.48708  66.48707  66.48705  66.48704  66.48703
#>  [92]  66.48702  66.48702  66.48701  66.48700  66.48699  66.48699  66.48698
#>  [99]  66.48698  66.48697
#> 
#> $linear.predictors
#>      5.6491   5.2684   4.9133   4.5822   4.2734   3.9853   3.7167   3.4662
#> 1  19.94286 20.25230 20.54089 20.81004 21.06104 21.29512 21.51343 21.71703
#> 2  19.94286 20.42306 20.87090 21.28855 21.67806 22.04131 22.38009 22.69603
#> 3  19.94286 19.78557 19.63889 19.50209 19.37451 19.25553 19.14457 19.04109
#> 4  19.94286 19.77419 19.61689 19.47019 19.33338 19.20579 19.08679 18.97582
#> 5  19.94286 19.95064 19.95789 19.96466 19.97097 19.97685 19.98234 19.98746
#> 6  19.94286 19.78557 19.63889 19.50209 19.37451 19.25553 19.14457 19.04109
#> 7  19.94286 19.59205 19.26489 18.95977 18.67522 18.40985 18.16236 17.93155
#> 8  19.94286 18.75535 17.64787 16.61504 15.65182 14.75351 13.91575 13.13445
#> 9  19.94286 18.65631 17.45647 16.33750 15.29394 14.32072 13.41309 12.56662
#> 10 19.94286 20.49136 21.00290 21.47996 21.92487 22.33979 22.72675 23.08763
#> 11 19.94286 20.82433 21.64640 22.41307 23.12806 23.79486 24.41673 24.99668
#> 12 19.94286 19.74004 19.55089 19.37449 19.20997 19.05655 18.91346 18.78002
#> 13 19.94286 19.78842 19.64439 19.51007 19.38480 19.26797 19.15901 19.05740
#> 14 19.94286 19.55790 19.19889 18.86407 18.55182 18.26061 17.98903 17.73575
#> 15 19.94286 19.55505 19.19339 18.85609 18.54153 18.24817 17.97459 17.71944
#> 16 19.94286 20.64219 21.29440 21.90265 22.46990 22.99893 23.49229 23.95241
#> 17 19.94286 20.52551 21.06890 21.57566 22.04827 22.48903 22.90008 23.28343
#> 18 19.94286 20.88239 21.75860 22.57576 23.33785 24.04857 24.71139 25.32954
#> 19 19.94286 19.93925 19.93589 19.93276 19.92983 19.92710 19.92456 19.92219
#> 20 19.94286 20.16693 20.37589 20.57078 20.75253 20.92203 21.08010 21.22752
#> 21 19.94286 19.71158 19.49589 19.29473 19.10714 18.93218 18.76902 18.61685
#>      3.2326   3.0148   2.8116    2.6221    2.4454    2.2806    2.1269    1.9835
#> 1  21.90690 22.08398 22.24912 22.403130 22.546763 22.680715 22.805639 22.922143
#> 2  22.99067 23.26546 23.52173 23.760730 23.983620 24.191488 24.385347 24.566139
#> 3  18.94458 18.85458 18.77064 18.692357 18.619352 18.551267 18.487771 18.428554
#> 4  18.87233 18.77581 18.68580 18.601850 18.523561 18.450549 18.382457 18.318954
#> 5  19.99223 19.99668 20.00083 20.004703 20.008314 20.011681 20.014822 20.017750
#> 6  18.94458 18.85458 18.77064 18.692357 18.619352 18.551267 18.487771 18.428554
#> 7  17.71630 17.51556 17.32834 17.153743 16.990913 16.839057 16.697435 16.565359
#> 8  12.40580 11.72627 11.09253 10.501503  9.950310  9.436266  8.956867  8.509778
#> 9  11.77721 11.04100 10.35441  9.714094  9.116933  8.560017  8.040636  7.556260
#> 10 23.42418 23.73806 24.03078 24.303770 24.558363 24.795798 25.017230 25.223738
#> 11 25.53755 26.04196 26.51238 26.951091 27.360236 27.741806 28.097660 28.429530
#> 12 18.65557 18.53951 18.43127 18.330330 18.236189 18.148394 18.066515 17.990155
#> 13 18.96264 18.87427 18.79185 18.714983 18.643299 18.576446 18.514099 18.455954
#> 14 17.49955 17.27926 17.07382 16.882223 16.703541 16.536902 16.381494 16.236559
#> 15 17.48148 17.25957 17.05261 16.859596 16.679594 16.511722 16.355165 16.209159
#> 16 24.38152 24.78171 25.15492 25.502984 25.827588 26.130314 26.412638 26.675934
#> 17 23.64094 23.97436 24.28530 24.575290 24.845735 25.097952 25.333171 25.552537
#> 18 25.90603 26.44367 26.94507 27.412675 27.848768 28.255469 28.634760 28.988489
#> 19 19.91998 19.91791 19.91599 19.914197 19.912523 19.910963 19.909508 19.908151
#> 20 21.36501 21.49323 21.61281 21.724330 21.828334 21.925328 22.015785 22.100145
#> 21 18.47494 18.34260 18.21917 18.104063 17.996713 17.896598 17.803231 17.716156
#>       1.8498    1.7252    1.6089    1.5005    1.3993    1.3050    1.2171
#> 1  23.003806 23.003806 23.003806 22.904699 22.807197 22.716266 22.631464
#> 2  24.692864 24.692864 24.692864 24.767711 24.841347 24.910020 24.974065
#> 3  18.387047 18.387047 18.387047 18.344900 18.303435 18.264764 18.228700
#> 4  18.274443 18.274443 18.274443 18.497255 18.716461 18.920893 19.111547
#> 5  20.019803 20.019803 20.019803 20.460280 20.893628 21.297769 21.674673
#> 6  18.387047 18.387047 18.387047 18.450780 18.513481 18.571956 18.626491
#> 7  16.472781 16.472781 16.472781 16.516109 16.558735 16.598488 16.635562
#> 8   8.196397  8.196397  8.196397  8.257138  8.316896  8.372627  8.424601
#> 9   7.216743  7.216743  7.216743  7.266506  7.315464  7.361122  7.403703
#> 10 25.368487 25.368487 25.368487 25.512916 25.655007 25.787522 25.911106
#> 11 28.662150 28.662150 28.662150 28.720408 28.777723 28.831175 28.881025
#> 12 17.936632 17.936632 17.936632 17.883114 17.830463 17.781361 17.735568
#> 13 18.415198 18.415198 18.415198 18.396147 18.377405 18.359925 18.343624
#> 14 16.134970 16.134970 16.134970 15.964834 15.797452 15.641351 15.495771
#> 15 16.106819 16.106819 16.106819 16.072406 16.038551 16.006978 15.977532
#> 16 26.860488 26.860488 26.860488 26.954330 27.046653 27.132754 27.213052
#> 17 25.706299 25.706299 25.706299 25.620818 25.536722 25.458293 25.385150
#> 18 29.236430 29.236430 29.236430 29.159360 29.083538 29.012826 28.946880
#> 19 19.907199 19.907199 19.907199 19.653098 19.403111 19.169971 18.952545
#> 20 22.159277 22.159277 22.159277 21.982706 21.808994 21.646989 21.495902
#> 21 17.655122 17.655122 17.655122 17.414482 17.177736 16.956947 16.751038
#>       1.1350    1.0585    0.9872    0.9207    0.8586    0.8007    0.7468
#> 1  22.552377 22.478620 22.409834 22.345685 22.285858 22.230064 22.178030
#> 2  25.033794 25.089497 25.141445 25.189893 25.235075 25.277213 25.316510
#> 3  18.195067 18.163700 18.134447 18.107166 18.081724 18.057996 18.035867
#> 4  19.289351 19.455172 19.609818 19.754040 19.888543 20.013980 20.130964
#> 5  22.026174 22.353986 22.659704 22.944817 23.210715 23.458693 23.689957
#> 6  18.677349 18.724781 18.769015 18.810268 18.848741 18.884621 18.918083
#> 7  16.670138 16.702383 16.732455 16.760500 16.786655 16.811047 16.833795
#> 8   8.473073  8.518277  8.560435  8.599752  8.636419  8.670615  8.702506
#> 9   7.443414  7.480449  7.514988  7.547199  7.577239  7.605254  7.631381
#> 10 26.026360 26.133847 26.234090 26.327576 26.414762 26.496072 26.571902
#> 11 28.927515 28.970871 29.011306 29.049015 29.084183 29.116981 29.147568
#> 12 17.692861 17.653032 17.615888 17.581247 17.548941 17.518812 17.490714
#> 13 18.328422 18.314244 18.301021 18.288690 18.277190 18.266465 18.256462
#> 14 15.360002 15.233384 15.115299 15.005173 14.902469 14.806687 14.717361
#> 15 15.950071 15.924461 15.900577 15.878303 15.857529 15.838156 15.820089
#> 16 27.287938 27.357777 27.422909 27.483651 27.540299 27.593130 27.642400
#> 17 25.316937 25.253321 25.193993 25.138663 25.087062 25.038939 24.994059
#> 18 28.885379 28.828022 28.774531 28.724645 28.678122 28.634734 28.594270
#> 19 18.749772 18.560666 18.384304 18.219829 18.066439 17.923387 17.789976
#> 20 21.354999 21.223592 21.101041 20.986750 20.880161 20.780757 20.688051
#> 21 16.559007 16.379918 16.212899 16.057137 15.911872 15.776398 15.650054
#>       0.6964    0.6495    0.6057    0.5649    0.5268    0.4913    0.4582
#> 1  22.129503 22.084246 22.042040 22.002678 21.976896 21.976895 21.976895
#> 2  25.353159 25.387338 25.419213 25.448940 25.468411 25.468412 25.468412
#> 3  18.015230 17.995984 17.978035 17.961295 17.950330 17.950330 17.950330
#> 4  20.240063 20.341809 20.436698 20.525192 20.583156 20.583159 20.583159
#> 5  23.905635 24.106777 24.294363 24.469306 24.583896 24.583902 24.583902
#> 6  18.949289 18.978393 19.005534 19.030847 19.047427 19.047428 19.047428
#> 7  16.855011 16.874796 16.893248 16.910456 16.921728 16.921728 16.921728
#> 8   8.732248  8.759985  8.785853  8.809977  8.825779  8.825780  8.825780
#> 9   7.655748  7.678472  7.699665  7.719429  7.732375  7.732376  7.732376
#> 10 26.642621 26.708574 26.770082 26.827444 26.865018 26.865019 26.865019
#> 11 29.176094 29.202697 29.227508 29.250646 29.265801 29.265802 29.265802
#> 12 17.464509 17.440071 17.417279 17.396024 17.382102 17.382101 17.382101
#> 13 18.247134 18.238435 18.230322 18.222755 18.217799 18.217799 18.217799
#> 14 14.634054 14.556363 14.483907 14.416335 14.372074 14.372072 14.372072
#> 15 15.803239 15.787525 15.772870 15.759202 15.750250 15.750250 15.750250
#> 16 27.688349 27.731202 27.771166 27.808437 27.832850 27.832851 27.832851
#> 17 24.952204 24.913170 24.876767 24.842817 24.820579 24.820578 24.820578
#> 18 28.556533 28.521339 28.488518 28.457908 28.437859 28.437858 28.437858
#> 19 17.665556 17.549522 17.441309 17.340388 17.274284 17.274281 17.274281
#> 20 20.601594 20.520964 20.445768 20.375640 20.329705 20.329703 20.329703
#> 21 15.532225 15.422338 15.319857 15.224282 15.161680 15.161677 15.161677
#>       0.4273    0.3985    0.3717    0.3466    0.3233    0.3015    0.2812
#> 1  21.976895 21.976895 21.976895 21.976895 21.976895 21.845507 21.813782
#> 2  25.468412 25.468412 25.468412 25.468412 25.468412 26.708923 26.731211
#> 3  17.950330 17.950330 17.950330 17.950330 17.950330 16.821870 16.864976
#> 4  20.583159 20.583159 20.583159 20.583159 20.583159 19.961711 20.007561
#> 5  24.583902 24.583902 24.583902 24.583902 24.583902 23.578699 23.575693
#> 6  19.047428 19.047428 19.047428 19.047428 19.047428 19.069212 19.037904
#> 7  16.921728 16.921728 16.921728 16.921728 16.921728 15.350652 15.368215
#> 8   8.825780  8.825780  8.825780  8.825780  8.825780 10.053563 10.046159
#> 9   7.732376  7.732376  7.732376  7.732376  7.732376  9.464447  9.457131
#> 10 26.865019 26.865019 26.865019 26.865019 26.865019 27.115210 27.137437
#> 11 29.265802 29.265802 29.265802 29.265802 29.265802 28.775788 28.772002
#> 12 17.382101 17.382101 17.382101 17.382101 17.382101 16.551012 16.594159
#> 13 18.217799 18.217799 18.217799 18.217799 18.217799 16.838798 16.881902
#> 14 14.372072 14.372072 14.372072 14.372072 14.372072 14.827437 14.819315
#> 15 15.750250 15.750250 15.750250 15.750250 15.750250 15.450651 15.493963
#> 16 27.832851 27.832851 27.832851 27.832851 27.832851 28.012427 28.034519
#> 17 24.820578 24.820578 24.820578 24.820578 24.820578 26.998282 26.994763
#> 18 28.437858 28.437858 28.437858 28.437858 28.437858 29.121132 29.117294
#> 19 17.274281 17.274281 17.274281 17.274281 17.274281 17.095873 17.087411
#> 20 20.329703 20.329703 20.329703 20.329703 20.329703 20.697506 20.614426
#> 21 15.161677 15.161677 15.161677 15.161677 15.161677 14.461298 14.350176
#>       0.2622    0.2445    0.2281    0.2127    0.1984   0.1850    0.1725
#> 1  21.783883 21.755980 21.741280 21.741252 21.741246 21.74124 21.741245
#> 2  26.751970 26.771319 26.781511 26.781529 26.781533 26.78153 26.781533
#> 3  16.906244 16.944823 16.965151 16.965190 16.965200 16.96520 16.965203
#> 4  20.050770 20.091095 20.112339 20.112378 20.112387 20.11239 20.112389
#> 5  23.572111 23.568691 23.566884 23.566878 23.566876 23.56687 23.566874
#> 6  19.008025 18.980102 18.965389 18.965361 18.965354 18.96535 18.965352
#> 7  15.385129 15.400950 15.409287 15.409304 15.409308 15.40931 15.409310
#> 8  10.038746 10.031783 10.028112 10.028104 10.028101 10.02810 10.028100
#> 9   9.449722  9.442755  9.439081  9.439073  9.439071  9.43907  9.439069
#> 10 27.158193 27.177545 27.187739 27.187757 27.187761 27.18776 27.187762
#> 11 28.768382 28.764998 28.763215 28.763211 28.763210 28.76321 28.763210
#> 12 16.635428 16.674005 16.694332 16.694372 16.694381 16.69438 16.694385
#> 13 16.923170 16.961749 16.982077 16.982117 16.982126 16.98213 16.982130
#> 14 14.811868 14.804938 14.801289 14.801283 14.801281 14.80128 14.801281
#> 15 15.535241 15.573810 15.594132 15.594171 15.594180 15.59418 15.594183
#> 16 28.055269 28.074627 28.084825 28.084844 28.084848 28.08485 28.084849
#> 17 26.991155 26.987759 26.985968 26.985964 26.985962 26.98596 26.985962
#> 18 29.113671 29.110290 29.108508 29.108505 29.108504 29.10850 29.108504
#> 19 17.079947 17.073033 17.069394 17.069389 17.069388 17.06939 17.069389
#> 20 20.535805 20.462399 20.423726 20.423653 20.423636 20.42363 20.423631
#> 21 14.245272 14.147350 14.095763 14.095667 14.095645 14.09564 14.095639
#>       0.1609    0.1500    0.1399    0.1305    0.1217    0.1135    0.1059
#> 1  21.741244 21.741244 21.741244 21.741244 21.741244 22.073773 22.073760
#> 2  26.781533 26.781533 26.781533 26.781533 26.781533 25.709586 25.709469
#> 3  16.965204 16.965204 16.965204 16.965204 16.965204 17.808456 17.808584
#> 4  20.112389 20.112389 20.112389 20.112389 20.112389 19.828129 19.828013
#> 5  23.566874 23.566874 23.566874 23.566874 23.566873 24.083787 24.083982
#> 6  18.965352 18.965351 18.965351 18.965351 18.965351 18.810352 18.810122
#> 7  15.409310 15.409310 15.409310 15.409310 15.409310 13.895860 13.895658
#> 8  10.028100 10.028100 10.028100 10.028100 10.028100 11.184700 11.184860
#> 9   9.439069  9.439069  9.439069  9.439069  9.439069  9.386719  9.386725
#> 10 27.187762 27.187762 27.187761 27.187761 27.187761 26.538362 26.538268
#> 11 28.763210 28.763210 28.763210 28.763210 28.763210 30.703720 30.703898
#> 12 16.694385 16.694385 16.694385 16.694385 16.694385 16.480869 16.480781
#> 13 16.982130 16.982130 16.982130 16.982130 16.982130 16.740436 16.740360
#> 14 14.801282 14.801282 14.801282 14.801282 14.801282 13.899398 13.899316
#> 15 15.594184 15.594184 15.594184 15.594184 15.594184 16.600183 16.600325
#> 16 28.084849 28.084849 28.084849 28.084849 28.084849 27.957857 27.957824
#> 17 26.985962 26.985961 26.985961 26.985961 26.985961 25.993860 25.993874
#> 18 29.108504 29.108504 29.108504 29.108504 29.108504 29.770740 29.770839
#> 19 17.069389 17.069389 17.069389 17.069389 17.069389 17.259327 17.259431
#> 20 20.423631 20.423631 20.423631 20.423631 20.423631 19.970507 19.970434
#> 21 14.095639 14.095638 14.095638 14.095638 14.095638 14.103382 14.103479
#>       0.0987    0.0921    0.0859    0.0801    0.0747    0.0696    0.0650
#> 1  22.073747 22.073735 22.073723 22.073711 22.073700 22.073689 22.265078
#> 2  25.709357 25.709250 25.709148 25.709052 25.708960 25.708872 25.858975
#> 3  17.808706 17.808823 17.808934 17.809039 17.809140 17.809236 17.699041
#> 4  19.827902 19.827796 19.827695 19.827598 19.827506 19.827418 19.867349
#> 5  24.084168 24.084347 24.084517 24.084679 24.084835 24.084983 23.906611
#> 6  18.809902 18.809693 18.809492 18.809301 18.809118 18.808943 18.545758
#> 7  13.895465 13.895282 13.895106 13.894939 13.894779 13.894627 13.949057
#> 8  11.185012 11.185157 11.185296 11.185429 11.185556 11.185676 11.295652
#> 9   9.386731  9.386736  9.386742  9.386748  9.386753  9.386759  9.508652
#> 10 26.538179 26.538094 26.538013 26.537936 26.537862 26.537792 26.801869
#> 11 30.704068 30.704230 30.704385 30.704531 30.704671 30.704805 30.746539
#> 12 16.480697 16.480616 16.480539 16.480465 16.480394 16.480327 16.427292
#> 13 16.740288 16.740219 16.740152 16.740089 16.740028 16.739970 16.723299
#> 14 13.899239 13.899165 13.899095 13.899028 13.898964 13.898903 13.541535
#> 15 16.600461 16.600590 16.600713 16.600831 16.600943 16.601050 16.539070
#> 16 27.957792 27.957762 27.957733 27.957705 27.957679 27.957654 28.118409
#> 17 25.993887 25.993901 25.993914 25.993927 25.993941 25.993954 25.807873
#> 18 29.770933 29.771023 29.771109 29.771191 29.771269 29.771343 29.552200
#> 19 17.259531 17.259625 17.259716 17.259802 17.259884 17.259962 17.499830
#> 20 19.970364 19.970297 19.970234 19.970173 19.970115 19.970060 19.942409
#> 21 14.103572 14.103661 14.103746 14.103826 14.103904 14.103977 14.203502
#>       0.0606    0.0565    0.0527   0.0491    0.0458   0.0427    0.0399   0.0372
#> 1  22.265206 22.265325 22.265436 22.26554 22.265633 22.26572 22.265804 22.26588
#> 2  25.859142 25.859299 25.859447 25.85958 25.859715 25.85984 25.859951 25.86006
#> 3  17.698916 17.698797 17.698685 17.69858 17.698478 17.69838 17.698292 17.69821
#> 4  19.867520 19.867679 19.867828 19.86797 19.868097 19.86822 19.868332 19.86844
#> 5  23.906295 23.906001 23.905725 23.90547 23.905227 23.90500 23.904790 23.90459
#> 6  18.545770 18.545786 18.545805 18.54583 18.545850 18.54588 18.545902 18.54593
#> 7  13.949308 13.949542 13.949761 13.94997 13.950159 13.95034 13.950509 13.95067
#> 8  11.295545 11.295444 11.295349 11.29526 11.295173 11.29509 11.295015 11.29494
#> 9   9.508656  9.508661  9.508665  9.50867  9.508675  9.50868  9.508685  9.50869
#> 10 26.802130 26.802373 26.802598 26.80281 26.803004 26.80319 26.803356 26.80351
#> 11 30.746458 30.746379 30.746303 30.74623 30.746160 30.74609 30.746026 30.74596
#> 12 16.427369 16.427441 16.427509 16.42757 16.427632 16.42769 16.427742 16.42779
#> 13 16.723408 16.723510 16.723605 16.72369 16.723776 16.72385 16.723925 16.72399
#> 14 13.541332 13.541147 13.540977 13.54082 13.540680 13.54055 13.540431 13.54032
#> 15 16.538936 16.538811 16.538692 16.53858 16.538473 16.53837 16.538277 16.53819
#> 16 28.118557 28.118694 28.118821 28.11894 28.119049 28.11915 28.119245 28.11933
#> 17 25.807569 25.807290 25.807033 25.80680 25.806580 25.80638 25.806195 25.80602
#> 18 29.551961 29.551740 29.551533 29.55134 29.551161 29.55099 29.550838 29.55069
#> 19 17.499981 17.500117 17.500240 17.50035 17.500452 17.50054 17.500625 17.50070
#> 20 19.942399 19.942391 19.942385 19.94238 19.942380 19.94238 19.942382 19.94238
#> 21 14.203541 14.203574 14.203602 14.20363 14.203646 14.20366 14.203677 14.20369
#>       0.0347    0.0323    0.0301    0.0281    0.0262    0.0245    0.0228
#> 1  22.265952 22.332591 22.332547 22.332505 22.332466 22.332429 22.332393
#> 2  25.860160 25.935023 25.935163 25.935297 25.935425 25.935548 25.935665
#> 3  17.698124 17.729402 17.729236 17.729074 17.728917 17.728763 17.728613
#> 4  19.868538 19.846787 19.846855 19.846925 19.846995 19.847065 19.847136
#> 5  23.904407 23.845876 23.845791 23.845706 23.845619 23.845532 23.845445
#> 6  18.545958 18.394635 18.394815 18.394995 18.395175 18.395353 18.395530
#> 7  13.950818 13.969742 13.969908 13.970070 13.970228 13.970383 13.970534
#> 8  11.294873 11.310454 11.310320 11.310190 11.310062 11.309938 11.309818
#> 9   9.508695  9.518945  9.518987  9.519028  9.519066  9.519103  9.519137
#> 10 26.803663 26.827198 26.827282 26.827366 26.827449 26.827531 26.827612
#> 11 30.745903 30.872767 30.872496 30.872231 30.871973 30.871723 30.871478
#> 12 16.427840 16.390300 16.390337 16.390376 16.390416 16.390458 16.390500
#> 13 16.724055 16.688676 16.688696 16.688719 16.688744 16.688770 16.688799
#> 14 13.540220 13.769162 13.769235 13.769296 13.769345 13.769384 13.769412
#> 15 16.538101 16.535617 16.535458 16.535305 16.535156 16.535011 16.534871
#> 16 28.119414 28.178075 28.178082 28.178090 28.178098 28.178106 28.178113
#> 17 25.805868 25.707250 25.707418 25.707576 25.707726 25.707868 25.708002
#> 18 29.550556 29.496798 29.496732 29.496668 29.496604 29.496541 29.496479
#> 19 17.500766 17.282739 17.282673 17.282617 17.282570 17.282531 17.282500
#> 20 19.942389 19.898981 19.899079 19.899174 19.899266 19.899356 19.899443
#> 21 14.203697 14.268984 14.268888 14.268794 14.268701 14.268608 14.268517
#>      0.0213    0.0198    0.0185    0.0173    0.0161    0.0150    0.0140
#> 1  22.33236 22.332329 22.332300 22.329598 22.329616 22.329634 22.329652
#> 2  25.93578 25.935885 25.935986 25.927662 25.927625 25.927588 25.927551
#> 3  17.72847 17.728326 17.728189 17.713633 17.713537 17.713447 17.713361
#> 4  19.84721 19.847278 19.847340 19.798951 19.799009 19.799066 19.799121
#> 5  23.84536 23.845271 23.845185 23.859188 23.859066 23.858949 23.858836
#> 6  18.39571 18.395881 18.396055 18.440543 18.440834 18.441113 18.441381
#> 7  13.97068 13.970824 13.970963 13.984978 13.985084 13.985183 13.985277
#> 8  11.30970 11.309586 11.309475 11.292074 11.291989 11.291909 11.291833
#> 9   9.51917  9.519201  9.519231  9.525918  9.525907  9.525895  9.525883
#> 10 26.82769 26.827770 26.827849 26.842797 26.842900 26.842998 26.843093
#> 11 30.87124 30.871010 30.870799 30.922009 30.921954 30.921903 30.921857
#> 12 16.39054 16.390588 16.390626 16.354802 16.354862 16.354921 16.354978
#> 13 16.68883 16.688861 16.688893 16.693478 16.693595 16.693709 16.693819
#> 14 13.76943 13.769442 13.769457 13.819507 13.819235 13.818970 13.818710
#> 15 16.53474 16.534604 16.534477 16.517879 16.517813 16.517751 16.517694
#> 16 28.17812 28.178128 28.178137 28.179128 28.179137 28.179146 28.179155
#> 17 25.70813 25.708247 25.708360 25.731756 25.731697 25.731637 25.731576
#> 18 29.49642 29.496359 29.496291 29.427616 29.427474 29.427338 29.427209
#> 19 17.28248 17.282461 17.282456 17.339395 17.339723 17.340041 17.340349
#> 20 19.89953 19.899611 19.899684 19.859146 19.859136 19.859126 19.859115
#> 21 14.26843 14.268338 14.268250 14.239942 14.239805 14.239675 14.239550
#>       0.0131    0.0122    0.0114    0.0106    0.0099    0.0092    0.0086
#> 1  22.329670 22.329688 22.329707 22.329725 22.329743 22.329761 22.329780
#> 2  25.927513 25.927475 25.927436 25.927398 25.927359 25.927320 25.927281
#> 3  17.713281 17.713205 17.713134 17.713067 17.713005 17.712946 17.712892
#> 4  19.799175 19.799227 19.799278 19.799327 19.799375 19.799421 19.799466
#> 5  23.858727 23.858623 23.858522 23.858425 23.858332 23.858242 23.858156
#> 6  18.441637 18.441883 18.442117 18.442341 18.442556 18.442760 18.442956
#> 7  13.985366 13.985450 13.985529 13.985604 13.985674 13.985740 13.985802
#> 8  11.291762 11.291695 11.291631 11.291572 11.291516 11.291464 11.291415
#> 9   9.525871  9.525858  9.525845  9.525832  9.525818  9.525805  9.525791
#> 10 26.843183 26.843269 26.843351 26.843430 26.843505 26.843577 26.843645
#> 11 30.921814 30.921775 30.921739 30.921707 30.921678 30.921652 30.921629
#> 12 16.355033 16.355087 16.355140 16.355191 16.355241 16.355289 16.355336
#> 13 16.693925 16.694028 16.694127 16.694223 16.694315 16.694404 16.694490
#> 14 13.818456 13.818207 13.817965 13.817728 13.817497 13.817272 13.817052
#> 15 16.517641 16.517591 16.517546 16.517504 16.517465 16.517429 16.517397
#> 16 28.179163 28.179171 28.179178 28.179186 28.179193 28.179200 28.179206
#> 17 25.731514 25.731451 25.731388 25.731324 25.731260 25.731195 25.731131
#> 18 29.427086 29.426968 29.426856 29.426750 29.426649 29.426552 29.426461
#> 19 17.340648 17.340938 17.341219 17.341491 17.341754 17.342009 17.342255
#> 20 19.859105 19.859094 19.859083 19.859072 19.859060 19.859049 19.859038
#> 21 14.239430 14.239317 14.239208 14.239104 14.239006 14.238912 14.238822
#>       0.0080    0.0075   0.0070    0.0065    0.0061    0.0056
#> 1  22.329798 22.329815 22.32983 22.329851 22.329868 22.329885
#> 2  25.927242 25.927203 25.92716 25.927126 25.927088 25.927050
#> 3  17.712841 17.712793 17.71275 17.712708 17.712670 17.712635
#> 4  19.799510 19.799552 19.79959 19.799632 19.799670 19.799707
#> 5  23.858073 23.857994 23.85792 23.857845 23.857776 23.857709
#> 6  18.443142 18.443320 18.44349 18.443651 18.443804 18.443950
#> 7  13.985860 13.985914 13.98596 13.986012 13.986056 13.986097
#> 8  11.291369 11.291326 11.29129 11.291250 11.291216 11.291184
#> 9   9.525778  9.525764  9.52575  9.525737  9.525723  9.525709
#> 10 26.843709 26.843771 26.84383 26.843885 26.843938 26.843988
#> 11 30.921608 30.921590 30.92157 30.921561 30.921549 30.921540
#> 12 16.355382 16.355426 16.35547 16.355511 16.355551 16.355590
#> 13 16.694573 16.694653 16.69473 16.694804 16.694875 16.694944
#> 14 13.816838 13.816630 13.81643 13.816230 13.816039 13.815853
#> 15 16.517367 16.517340 16.51732 16.517294 16.517275 16.517258
#> 16 28.179212 28.179218 28.17922 28.179229 28.179234 28.179239
#> 17 25.731067 25.731003 25.73094 25.730875 25.730812 25.730749
#> 18 29.426374 29.426292 29.42621 29.426141 29.426072 29.426006
#> 19 17.342493 17.342723 17.34294 17.343159 17.343365 17.343564
#> 20 19.859026 19.859015 19.85900 19.858993 19.858982 19.858971
#> 21 14.238737 14.238656 14.23858 14.238507 14.238438 14.238372
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 22.8 18.7 18.1 22.8 19.2 15.2 10.4 10.4 32.4 30.4 15.5 15.2 13.3 19.2
#> [16] 27.3 26.0 30.4 15.8 19.7 15.0
#> 
#> $X
#>               am        carb        cyl        disp        drat       gear
#>  [1,]  1.0488088  0.62831212 -0.1608169 -0.60524443  0.49991005  0.2399041
#>  [2,]  1.0488088 -1.09271673 -1.2865350 -1.01146871  0.41257135  0.2399041
#>  [3,] -0.9534626 -0.51904045  0.9649013  0.95715668 -0.81017036 -1.0195922
#>  [4,] -0.9534626 -1.09271673 -0.1608169 -0.09746407 -1.49141217 -1.0195922
#>  [5,] -0.9534626 -0.51904045 -1.2865350 -0.75523493  0.53484552  0.2399041
#>  [6,] -0.9534626  0.62831212 -0.1608169 -0.54587319  0.53484552  0.2399041
#>  [7,] -0.9534626  0.05463584  0.9649013  0.29938581 -0.94991227 -1.0195922
#>  [8,] -0.9534626  0.62831212  0.9649013  1.83210129 -1.19446061 -1.0195922
#>  [9,] -0.9534626  0.62831212  0.9649013  1.73835723 -1.07218644 -1.0195922
#> [10,]  1.0488088 -1.09271673 -1.2865350 -1.24036048  0.81432934  0.2399041
#> [11,]  1.0488088 -0.51904045 -1.2865350 -1.26379649  2.29908714  0.2399041
#> [12,] -0.9534626 -0.51904045  0.9649013  0.62905244 -1.49141217 -1.0195922
#> [13,] -0.9534626 -0.51904045  0.9649013  0.51968437 -0.81017036 -1.0195922
#> [14,] -0.9534626  0.62831212  0.9649013  0.87903662  0.20295849 -1.0195922
#> [15,] -0.9534626 -0.51904045  0.9649013  1.26963690 -0.93244453 -1.0195922
#> [16,]  1.0488088 -1.09271673 -1.2865350 -1.23801687  0.81432934  0.2399041
#> [17,]  1.0488088 -0.51904045 -1.2865350 -0.91538105  1.42570020  1.4994004
#> [18,]  1.0488088 -0.51904045 -1.2865350 -1.11224359  0.27282944  1.4994004
#> [19,]  1.0488088  0.62831212  0.9649013  0.88684863  1.05887769  1.4994004
#> [20,]  1.0488088  1.77566469 -0.1608169 -0.72242451  0.01081336  1.4994004
#> [21,]  1.0488088  2.92301725  0.9649013  0.49624835 -0.12892855  1.4994004
#>                hp       qsec         vs           wt
#>  [1,] -0.58675556 -0.5907913 -0.7844645 -0.541845223
#>  [2,] -0.82670028  0.5505963  1.2747549 -0.840840135
#>  [3,]  0.33068011 -0.2934997 -0.7844645  0.275407535
#>  [4,] -0.65732754  1.4053098  1.2747549  0.295340530
#>  [5,] -0.79847149  2.8280628  1.2747549 -0.013620879
#>  [6,] -0.40326843  0.3860241  1.2747549  0.275407535
#>  [7,]  0.40125209  0.2267607 -0.7844645  0.614268435
#>  [8,]  0.75411196  0.2161432 -0.7844645  2.079343503
#>  [9,]  0.89525591  0.1312027 -0.7844645  2.252760552
#> [10,] -1.20778894  1.0071514  1.2747549 -0.960438100
#> [11,] -1.40539047  0.5028173  1.2747549 -1.543478178
#> [12,] -0.02217976 -0.3731314 -0.7844645  0.355139512
#> [13,] -0.02217976 -0.1448538 -0.7844645  0.270424287
#> [14,]  1.31868776 -1.1482132 -0.7844645  0.674067418
#> [15,]  0.33068011 -0.2775733 -0.7844645  0.679050666
#> [16,] -1.20778894  0.7045509  1.2747549 -1.224550272
#> [17,] -0.85492907 -0.4633806 -0.7844645 -1.020237082
#> [18,] -0.54441238 -0.3572050  1.2747549 -1.645136448
#> [19,]  1.58686126 -1.6313122 -0.7844645  0.006312115
#> [20,]  0.33068011 -1.1004342 -0.7844645 -0.392347768
#> [21,]  2.58898331 -1.5782244 -0.7844645  0.404971997
#> attr(,"center")
#>  [1]   0.4761905   2.9047619   6.2857143 237.4761905   3.6138095   3.8095238
#>  [7] 151.5714286  17.5728571   0.3809524   3.1636667
#> attr(,"scale")
#>  [1]   0.4994328   1.7431434   1.7766436 128.0081021   0.5724839   0.7939682
#>  [7]  70.8496540   1.8836721   0.4856209   1.0033616
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 37
#> 
#> $lambda.min
#> [1] 0.4582175
#> 
#> $null.dev
#> [1] 39.08721
#> 
#> $Bias
#> [1] 2.155866
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 9.669357