Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 28.037739 25.866174 23.208561 20.441921 18.164981 16.305239 14.800225
#>   [8] 13.596154 12.646756 11.912267 11.358544 10.956307 10.680475 10.509597
#>  [15] 10.425350 10.411101 10.163212  9.951088  9.861005  9.907615  9.956120
#>  [22]  9.994880 10.034359 10.075401 10.096952 10.124520 10.181867 10.296538
#>  [29] 10.377177 10.409213 10.503430 10.552615 10.228184  9.699610  9.867813
#>  [36]  9.729107  9.446549  9.448052  9.459063  9.555748  9.687235  9.699543
#>  [43] 10.670698 14.818689 14.800014 15.514755 15.423688 15.649213 15.620373
#>  [50] 15.709133 15.724187 15.701164 15.609714 15.136627 15.038467 14.565085
#>  [57] 15.088339 15.389270 16.541847 16.542295 34.368369 34.419308 34.422826
#>  [64] 35.178861 35.314457 35.480750 35.483132 35.485438 35.425096 35.244835
#>  [71] 35.320552 35.321844 35.322760 36.566052 36.523534 36.534897 36.360761
#>  [78] 36.361962 36.363174 36.364397 36.365630 36.366872 36.313223 36.684523
#>  [85] 36.685641 36.686779 36.687935 36.689104 36.690286 39.065100 39.065758
#>  [92] 39.074235 39.074892 39.075545 39.076190 39.076828 39.077457 39.078076
#>  [99] 39.078637 39.072661
#> 
#> $cvse
#>   [1]  8.979632  8.956631  8.685814  8.020577  7.417360  6.868261  6.367585
#>   [8]  5.911358  5.496924  5.122621  4.787488  4.491018  4.232914  4.012859
#>  [15]  3.830294  3.684312  3.529046  3.390692  3.320330  3.318321  3.316544
#>  [22]  3.316586  3.316841  3.317051  3.316446  3.315602  3.314277  3.310255
#>  [29]  3.292643  3.281221  3.272602  3.262866  3.078617  2.945449  2.809614
#>  [36]  2.760271  2.773518  2.773640  2.774146  2.760679  2.746461  2.746566
#>  [43]  2.768391  5.699914  5.701970  5.683985  5.690261  5.693789  5.694883
#>  [50]  5.682475  5.664907  5.669021  5.672881  5.696535  5.693487  5.692101
#>  [57]  4.990999  5.219842  5.605340  5.605136 24.050033 24.219956 24.222782
#>  [64] 24.206977 24.223541 24.222552 24.225145 24.227695 24.232729 24.240530
#>  [71] 24.238260 24.240654 24.243013 24.238780 24.244167 24.247073 24.261792
#>  [78] 24.263932 24.266036 24.268104 24.270138 24.272137 24.272893 24.260670
#>  [85] 24.262556 24.264411 24.266235 24.268029 24.269793 26.681187 26.682328
#>  [92] 26.683070 26.684139 26.685173 26.686174 26.687141 26.688076 26.688979
#>  [99] 26.689852 26.690938
#> 
#> $fold
#>  [1]  1  8  9  5  2  2  1  3  1  9  3  5  4 10  6  7  6  8  7 10  4
#> 
#> $lambda
#>   [1] 4.451725256 4.151693873 3.871883601 3.610931605 3.367566901 3.140604162
#>   [7] 2.928937953 2.731537338 2.547440863 2.375751875 2.215634150 2.066307824
#>  [13] 1.927045592 1.797169169 1.676045982 1.563086092 1.457739321 1.359492570
#>  [19] 1.267867321 1.182417307 1.102726338 1.028406273 0.959095132 0.894455330
#>  [25] 0.834172035 0.777951631 0.725520295 0.676622656 0.631020553 0.588491880
#>  [31] 0.548829497 0.511840225 0.477343907 0.445172526 0.415169387 0.387188360
#>  [37] 0.361093161 0.336756690 0.314060416 0.292893795 0.273153734 0.254744086
#>  [43] 0.237575188 0.221563415 0.206630782 0.192704559 0.179716917 0.167604598
#>  [49] 0.156308609 0.145773932 0.135949257 0.126786732 0.118241731 0.110272634
#>  [55] 0.102840627 0.095909513 0.089445533 0.083417203 0.077795163 0.072552030
#>  [61] 0.067662266 0.063102055 0.058849188 0.054882950 0.051184023 0.047734391
#>  [67] 0.044517253 0.041516939 0.038718836 0.036109316 0.033675669 0.031406042
#>  [73] 0.029289380 0.027315373 0.025474409 0.023757519 0.022156342 0.020663078
#>  [79] 0.019270456 0.017971692 0.016760460 0.015630861 0.014577393 0.013594926
#>  [85] 0.012678673 0.011824173 0.011027263 0.010284063 0.009590951 0.008944553
#>  [91] 0.008341720 0.007779516 0.007255203 0.006766227 0.006310206 0.005884919
#>  [97] 0.005488295 0.005118402 0.004773439 0.004451725
#> 
#> $fit
#> $beta
#>               4.4517     4.1517     3.8719     3.6109     3.3676    3.1406
#> (Intercept) 19.70476 20.9417540 22.3842873 23.7295986 24.9842404 26.154324
#> am           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> cyl          0.00000 -0.2061654 -0.4465876 -0.6708061 -0.8799131 -1.074927
#> disp         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> wt           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#>                2.9289    2.7315    2.5474    2.3758    2.2156    2.0663
#> (Intercept) 27.245547 28.263226 29.212316 30.097441 30.922912 31.692748
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.256798 -1.426411 -1.584592 -1.732113 -1.869692 -1.997998
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.9270    1.7972    1.6760    1.5631    1.4577    1.3595
#> (Intercept) 32.410701 33.080265 33.704703 34.287057 34.798512 34.798512
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.117656 -2.229251 -2.333324 -2.430382 -2.515625 -2.515625
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.2679    1.1824    1.1027    1.0284     0.9591     0.8945
#> (Intercept) 34.798512 34.798512 34.798512 34.798512 34.6961674 34.4639719
#> am           0.000000  0.000000  0.000000  0.000000  0.1032665  0.3375539
#> carb         0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> cyl         -2.515625 -2.515625 -2.515625 -2.515625 -2.5059438 -2.4839793
#> disp         0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#>                0.8342     0.7780     0.7255    0.6766    0.6310    0.5885
#> (Intercept) 34.247327 34.0452823 33.8568551 33.681127 33.517243 33.364404
#> am           0.556151  0.7600155  0.9501402  1.127451  1.292812  1.447028
#> carb         0.000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> cyl         -2.463486 -2.4443735 -2.4265494 -2.409926 -2.394424 -2.379966
#> disp         0.000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#>                0.5488    0.5118    0.4773    0.4452    0.4152    0.3872
#> (Intercept) 33.221866 36.820579 36.820413 36.820315 36.820257 36.820223
#> am           1.590850  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.366483 -1.613641 -1.613494 -1.613407 -1.613355 -1.613324
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000 -2.264350 -2.264569 -2.264699 -2.264776 -2.264821
#>                0.3611    0.3368      0.3141      0.2929     0.2732      0.2547
#> (Intercept) 36.820202 36.820190 32.56683779 32.49233327 32.4913610 32.49060523
#> am           0.000000  0.000000  1.84647361  1.89029881  1.8907065  1.89100684
#> carb         0.000000  0.000000  0.00000000  0.00000000  0.0000000  0.00000000
#> cyl         -1.613306 -1.613295 -0.87965040 -0.87786269 -0.8776712 -0.87750174
#> disp         0.000000  0.000000  0.00000000  0.00000000  0.0000000  0.00000000
#> drat         0.000000  0.000000  0.00000000  0.00000000  0.0000000  0.00000000
#> gear         0.000000  0.000000  0.00000000  0.00000000  0.0000000  0.00000000
#> hp           0.000000  0.000000 -0.02460805 -0.02478766 -0.0247931 -0.02479744
#> qsec         0.000000  0.000000  0.00000000  0.00000000  0.0000000  0.00000000
#> vs           0.000000  0.000000  0.00000000  0.00000000  0.0000000  0.00000000
#> wt          -2.264848 -2.264864 -1.45215989 -1.43043315 -1.4302972 -1.43022200
#>                 0.2376      0.2216     0.2066     0.1927      0.1797
#> (Intercept) 32.4900191 32.48956542 32.4892150 32.4889448 32.48873692
#> am           1.8912276  1.89138930  1.8915074  1.8915933  1.89165564
#> carb         0.0000000  0.00000000  0.0000000  0.0000000  0.00000000
#> cyl         -0.8773549 -0.87722985 -0.8771247 -0.8770372 -0.87696503
#> disp         0.0000000  0.00000000  0.0000000  0.0000000  0.00000000
#> drat         0.0000000  0.00000000  0.0000000  0.0000000  0.00000000
#> gear         0.0000000  0.00000000  0.0000000  0.0000000  0.00000000
#> hp          -0.0248009 -0.02480363 -0.0248058 -0.0248075 -0.02480884
#> qsec         0.0000000  0.00000000  0.0000000  0.0000000  0.00000000
#> vs           0.0000000  0.00000000  0.0000000  0.0000000  0.00000000
#> wt          -1.4301863 -1.43017558 -1.4301799 -1.4301927 -1.43020959
#>                  0.1676     0.1563      0.1458      0.1359      0.1268
#> (Intercept) 32.48857724 32.4884548 32.48836111 20.62424267 18.90524004
#> am           1.89170062  1.8917329  1.89175604  2.39734118  2.14284229
#> carb         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> cyl         -0.87690602 -0.8768581 -0.87681929 -0.42392258 -0.54606174
#> disp         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> drat         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.0000000  0.00000000  0.45663540  0.62029502
#> hp          -0.02480989 -0.0248107 -0.02481134 -0.02430795 -0.02311852
#> qsec         0.00000000  0.0000000  0.00000000  0.42659413  0.56988139
#> vs           0.00000000  0.0000000  0.00000000  0.00000000 -0.99735931
#> wt          -1.43022791 -1.4302460 -1.43026279 -1.60002126 -1.70331706
#>                  0.1182      0.1103     0.1028      0.0959      0.0894
#> (Intercept) 18.90750746 17.95993321 17.9642045 17.96749812 17.96996833
#> am           2.14266330  2.01811530  2.0178002  2.01756310  2.01739064
#> carb         0.00000000 -0.30922196 -0.3089960 -0.30879579 -0.30862051
#> cyl         -0.54632242 -0.46036581 -0.4605589 -0.46069273 -0.46077760
#> disp         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> drat         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> gear         0.62021892  1.00537686  1.0052410  1.00512450  1.00502641
#> hp          -0.02311625 -0.02126682 -0.0212729 -0.02127834 -0.02128317
#> qsec         0.56983356  0.52421177  0.5240597  0.52394209  0.52385320
#> vs          -0.99761010 -0.97581614 -0.9758685 -0.97589115 -0.97588987
#> wt          -1.70322284 -1.55375744 -1.5535970 -1.55347872 -1.55339506
#>                  0.0834      0.0778      0.0726      0.0677       0.0631
#> (Intercept) 17.97175211 17.97297012 17.97372769 17.97411599 12.200912966
#> am           2.01727124  2.01719488  2.01715295  2.01713811  2.140545038
#> carb        -0.30846872 -0.30833871 -0.30822860 -0.30813645 -0.743372054
#> cyl         -0.46082272 -0.46083607 -0.46082456 -0.46079408 -0.203744480
#> disp         0.00000000  0.00000000  0.00000000  0.00000000 -0.018549253
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.845899072
#> gear         1.00494537  1.00487981  1.00482808  1.00478849  0.988761178
#> hp          -0.02128741 -0.02129113 -0.02129435 -0.02129713 -0.007989475
#> qsec         0.52378810  0.52374249  0.52371268  0.52369548  0.501176498
#> vs          -0.97586970 -0.97583504 -0.97578970 -0.97573689 -1.467586869
#> wt          -1.55333953 -1.55330656 -1.55329140 -1.55329006  0.000000000
#>                   0.0588       0.0549       0.0512       0.0477       0.0445
#> (Intercept) 12.198177503 12.196358224 12.195293437 12.194842496 12.194883784
#> am           2.140926933  2.141228180  2.141460511  2.141634374  2.141759031
#> carb        -0.743558545 -0.743718205 -0.743853937 -0.743968447 -0.744064235
#> cyl         -0.203524668 -0.203362618 -0.203248983 -0.203175544 -0.203135116
#> disp        -0.018546660 -0.018544431 -0.018542528 -0.018540916 -0.018539563
#> drat         0.845836850  0.845743207  0.845626438  0.845493486  0.845350101
#> gear         0.989040128  0.989281297  0.989488428  0.989665076  0.989814576
#> hp          -0.007990423 -0.007991108 -0.007991573 -0.007991854 -0.007991983
#> qsec         0.501195222  0.501198429  0.501189347  0.501170735  0.501144933
#> vs          -1.467113320 -1.466717368 -1.466389612 -1.466121448 -1.465905044
#> wt           0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#>                  0.0415       0.0387       0.0361       0.0337       0.0314
#> (Intercept) 12.19531280 12.196040336 12.196990821 12.198100727 12.199317140
#> am           2.14184264  2.141892364  2.141914435  2.141914259  2.141896495
#> carb        -0.74414359 -0.744208623 -0.744261220 -0.744303100 -0.744335802
#> cyl         -0.20312145 -0.203129157 -0.203153599 -0.203190830 -0.203237510
#> disp        -0.01853844 -0.018537513 -0.018536762 -0.018536163 -0.018535695
#> drat         0.84520100  0.845050005  0.844900156  0.844753833  0.844612855
#> gear         0.98994004  0.990044331  0.990130096  0.990199737  0.990255434
#> hp          -0.00799199 -0.007991899 -0.007991731 -0.007991505 -0.007991237
#> qsec         0.50111391  0.501079316  0.501042505  0.501004590  0.500966468
#> vs          -1.46573332 -1.465599894 -1.465499077 -1.465425797 -1.465375569
#> wt           0.00000000  0.000000000  0.000000000  0.000000000  0.000000000
#>                   0.0293       0.0273       0.0255       0.0238       0.0222
#> (Intercept) 12.200596440 12.201903095 11.479135360 11.477965872 11.476828336
#> am           2.141865125  2.141823534  2.233956017  2.233992293  2.234027577
#> carb        -0.744360698 -0.744379005 -0.942284064 -0.942405223 -0.942523073
#> cyl         -0.203290839 -0.203348489 -0.182886738 -0.182806542 -0.182728537
#> disp        -0.018535338 -0.018535075 -0.025844328 -0.025848645 -0.025852844
#> drat         0.844478558  0.844351885  0.952977902  0.953093045  0.953205041
#> gear         0.990299151  0.990332646  1.107165562  1.107256565  1.107345081
#> hp          -0.007990939 -0.007990622 -0.003616995 -0.003614835 -0.003612734
#> qsec         0.500928848  0.500892282  0.424861653  0.424835126  0.424809324
#> vs          -1.465344449 -1.465328988 -1.597337479 -1.597394253 -1.597449475
#> wt           0.000000000  0.000000000  0.814456269  0.814877762  0.815287740
#>                  0.0207       0.0193       0.0180       0.0168       0.0156
#> (Intercept) 11.47572188 11.474645653 11.473598831 11.472580610 11.471590208
#> am           2.23406190  2.234095281  2.234127752  2.234159336  2.234190056
#> carb        -0.94263770 -0.942749200 -0.942857651 -0.942963139 -0.943065744
#> cyl         -0.18265266 -0.182578862 -0.182507078 -0.182437254 -0.182369339
#> disp        -0.02585693 -0.025860901 -0.025864765 -0.025868524 -0.025872180
#> drat         0.95331398  0.953419937  0.953523001  0.953623250  0.953720759
#> gear         1.10743118  1.107514924  1.107596382  1.107675614  1.107752681
#> hp          -0.00361069 -0.003608702 -0.003606768 -0.003604887 -0.003603057
#> qsec         0.42478423  0.424759815  0.424736071  0.424712975  0.424690511
#> vs          -1.59750319 -1.597555433 -1.597606251 -1.597655680 -1.597703758
#> wt           0.81568651  0.816074394  0.816451675  0.816818647  0.817175593
#>                   0.0146       0.0136       0.0127       0.0118       0.0110
#> (Intercept) 11.470626867 11.469689846 11.468778428 11.467891911 11.467029616
#> am           2.234219938  2.234249003  2.234277274  2.234304773  2.234331520
#> carb        -0.943165546 -0.943262621 -0.943357044 -0.943448887 -0.943538220
#> cyl         -0.182303278 -0.182239023 -0.182176524 -0.182115732 -0.182056601
#> disp        -0.025875736 -0.025879195 -0.025882559 -0.025885831 -0.025889014
#> drat         0.953815605  0.953907858  0.953997591  0.954084872  0.954169769
#> gear         1.107827642  1.107900555  1.107971476  1.108040460  1.108107558
#> hp          -0.003601278 -0.003599547 -0.003597863 -0.003596226 -0.003594633
#> qsec         0.424668660  0.424647407  0.424626734  0.424606626  0.424587067
#> vs          -1.597750522 -1.597796009 -1.597840252 -1.597883287 -1.597925146
#> wt           0.817522785  0.817860491  0.818188969  0.818508472  0.818819245
#>                   0.0103       0.0096       0.0089       0.0083       0.0078
#> (Intercept) 11.466190882 11.465375064 11.464581537 11.463809691 11.463058935
#> am           2.234357536  2.234382842  2.234407456  2.234431398  2.234454685
#> carb        -0.943625113 -0.943709631 -0.943791840 -0.943871803 -0.943949580
#> cyl         -0.181999085 -0.181943141 -0.181888726 -0.181835797 -0.181784315
#> disp        -0.025892110 -0.025895122 -0.025898051 -0.025900900 -0.025903671
#> drat         0.954252346  0.954332666  0.954410792  0.954486783  0.954560698
#> gear         1.108172823  1.108236305  1.108298053  1.108358113  1.108416532
#> hp          -0.003593084 -0.003591577 -0.003590111 -0.003588685 -0.003587298
#> qsec         0.424568043  0.424549539  0.424531540  0.424514033  0.424497005
#> vs          -1.597965861 -1.598005463 -1.598043983 -1.598081451 -1.598117894
#> wt           0.819121527  0.819415549  0.819701537  0.819979711  0.820250284
#>                   0.0073       0.0068       0.0063      0.0059       0.0055
#> (Intercept) 11.462328691 11.461618400 11.460927517 11.46025551 11.459601866
#> am           2.234477337  2.234499369  2.234520799  2.23454164  2.234561919
#> carb        -0.944025233 -0.944098819 -0.944170394 -0.94424001 -0.944307730
#> cyl         -0.181734239 -0.181685532 -0.181638155 -0.18159207 -0.181547249
#> disp        -0.025906367 -0.025908989 -0.025911539 -0.02591402 -0.025916432
#> drat         0.954632593  0.954702524  0.954770543  0.95483670  0.954901059
#> gear         1.108473355  1.108528626  1.108582386  1.10863468  1.108685540
#> hp          -0.003585949 -0.003584637 -0.003583361 -0.00358212 -0.003580912
#> qsec         0.424480442  0.424464331  0.424448660  0.42443342  0.424418592
#> vs          -1.598153342 -1.598187822 -1.598221359 -1.59825398 -1.598285709
#> wt           0.820513465  0.820769454  0.821018448  0.82126064  0.821496213
#>                   0.0051       0.0048       0.0045
#> (Intercept) 11.458966081 11.458347668 11.457746152
#> am           2.234581641  2.234600823  2.234619481
#> carb        -0.944373597 -0.944437664 -0.944499980
#> cyl         -0.181503651 -0.181461244 -0.181419995
#> disp        -0.025918779 -0.025921062 -0.025923282
#> drat         0.954963654  0.955024539  0.955083760
#> gear         1.108735013  1.108783134  1.108829941
#> hp          -0.003579738 -0.003578595 -0.003577484
#> qsec         0.424404172  0.424390145  0.424376502
#> vs          -1.598316572 -1.598346591 -1.598375790
#> wt           0.821725349  0.821948225  0.822165011
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   5   4   4   4   4   4   4   4   4  35   1   1   1   1
#>  [37]   1   1  70  10   1   1   1   1   1   1   1   1   1   1 129  88   1  44
#>  [55]   1   1   1   1   1   1   1 126   1   1   1   1   1   1   1   1   1   1
#>  [73]   1   1 153   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.451725256 4.151693873 3.871883601 3.610931605 3.367566901 3.140604162
#>   [7] 2.928937953 2.731537338 2.547440863 2.375751875 2.215634150 2.066307824
#>  [13] 1.927045592 1.797169169 1.676045982 1.563086092 1.457739321 1.359492570
#>  [19] 1.267867321 1.182417307 1.102726338 1.028406273 0.959095132 0.894455330
#>  [25] 0.834172035 0.777951631 0.725520295 0.676622656 0.631020553 0.588491880
#>  [31] 0.548829497 0.511840225 0.477343907 0.445172526 0.415169387 0.387188360
#>  [37] 0.361093161 0.336756690 0.314060416 0.292893795 0.273153734 0.254744086
#>  [43] 0.237575188 0.221563415 0.206630782 0.192704559 0.179716917 0.167604598
#>  [49] 0.156308609 0.145773932 0.135949257 0.126786732 0.118241731 0.110272634
#>  [55] 0.102840627 0.095909513 0.089445533 0.083417203 0.077795163 0.072552030
#>  [61] 0.067662266 0.063102055 0.058849188 0.054882950 0.051184023 0.047734391
#>  [67] 0.044517253 0.041516939 0.038718836 0.036109316 0.033675669 0.031406042
#>  [73] 0.029289380 0.027315373 0.025474409 0.023757519 0.022156342 0.020663078
#>  [79] 0.019270456 0.017971692 0.016760460 0.015630861 0.014577393 0.013594926
#>  [85] 0.012678673 0.011824173 0.011027263 0.010284063 0.009590951 0.008944553
#>  [91] 0.008341720 0.007779516 0.007255203 0.006766227 0.006310206 0.005884919
#>  [97] 0.005488295 0.005118402 0.004773439 0.004451725
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 31
#> 
#> $loss
#>   [1] 542.62952 478.96455 411.59252 355.42873 308.84932 270.45300 239.03128
#>   [8] 213.54273 193.09049 176.90290 164.31661 154.76193 147.75015 142.86256
#>  [15] 139.74086 138.07894 137.61390 137.61390 137.61390 137.61390 137.61390
#>  [22] 137.61390 135.54301 131.20690 127.61463 124.65895 122.24557 120.29325
#>  [29] 118.73206 117.50183 116.55084 101.66653 101.66653 101.66652 101.66652
#>  [36] 101.66652 101.66652 101.66652  85.16972  85.16506  85.16505  85.16505
#>  [43]  85.16505  85.16504  85.16504  85.16504  85.16504  85.16504  85.16504
#>  [50]  85.16504  81.75374  80.35798  80.35798  79.15057  79.15056  79.15055
#>  [57]  79.15055  79.15054  79.15054  79.15054  79.15054  75.02655  75.02654
#>  [64]  75.02653  75.02653  75.02653  75.02653  75.02652  75.02652  75.02652
#>  [71]  75.02652  75.02652  75.02652  75.02652  74.60641  74.60640  74.60639
#>  [78]  74.60638  74.60637  74.60636  74.60636  74.60635  74.60634  74.60634
#>  [85]  74.60633  74.60632  74.60632  74.60631  74.60631  74.60630  74.60630
#>  [92]  74.60630  74.60629  74.60629  74.60628  74.60628  74.60628  74.60628
#>  [99]  74.60627  74.60627
#> 
#> $linear.predictors
#>      4.4517   4.1517   3.8719   3.6109   3.3676   3.1406   2.9289   2.7315
#> 1  19.70476 19.70476 19.70476 19.70476 19.70476 19.70476 19.70476 19.70476
#> 2  19.70476 19.70476 19.70476 19.70476 19.70476 19.70476 19.70476 19.70476
#> 3  19.70476 20.11709 20.59794 21.04637 21.46459 21.85462 22.21836 22.55758
#> 4  19.70476 19.70476 19.70476 19.70476 19.70476 19.70476 19.70476 19.70476
#> 5  19.70476 20.11709 20.59794 21.04637 21.46459 21.85462 22.21836 22.55758
#> 6  19.70476 20.11709 20.59794 21.04637 21.46459 21.85462 22.21836 22.55758
#> 7  19.70476 19.70476 19.70476 19.70476 19.70476 19.70476 19.70476 19.70476
#> 8  19.70476 19.70476 19.70476 19.70476 19.70476 19.70476 19.70476 19.70476
#> 9  19.70476 19.29243 18.81159 18.36315 17.94494 17.55491 17.19117 16.85194
#> 10 19.70476 19.29243 18.81159 18.36315 17.94494 17.55491 17.19117 16.85194
#> 11 19.70476 19.29243 18.81159 18.36315 17.94494 17.55491 17.19117 16.85194
#> 12 19.70476 19.29243 18.81159 18.36315 17.94494 17.55491 17.19117 16.85194
#> 13 19.70476 19.29243 18.81159 18.36315 17.94494 17.55491 17.19117 16.85194
#> 14 19.70476 20.11709 20.59794 21.04637 21.46459 21.85462 22.21836 22.55758
#> 15 19.70476 20.11709 20.59794 21.04637 21.46459 21.85462 22.21836 22.55758
#> 16 19.70476 19.29243 18.81159 18.36315 17.94494 17.55491 17.19117 16.85194
#> 17 19.70476 20.11709 20.59794 21.04637 21.46459 21.85462 22.21836 22.55758
#> 18 19.70476 20.11709 20.59794 21.04637 21.46459 21.85462 22.21836 22.55758
#> 19 19.70476 19.29243 18.81159 18.36315 17.94494 17.55491 17.19117 16.85194
#> 20 19.70476 19.29243 18.81159 18.36315 17.94494 17.55491 17.19117 16.85194
#> 21 19.70476 20.11709 20.59794 21.04637 21.46459 21.85462 22.21836 22.55758
#>      2.5474   2.3758   2.2156   2.0663   1.9270   1.7972   1.6760   1.5631
#> 1  19.70476 19.70476 19.70476 19.70476 19.70476 19.70476 19.70476 19.70476
#> 2  19.70476 19.70476 19.70476 19.70476 19.70476 19.70476 19.70476 19.70476
#> 3  22.87395 23.16899 23.44415 23.70076 23.94007 24.16326 24.37141 24.56553
#> 4  19.70476 19.70476 19.70476 19.70476 19.70476 19.70476 19.70476 19.70476
#> 5  22.87395 23.16899 23.44415 23.70076 23.94007 24.16326 24.37141 24.56553
#> 6  22.87395 23.16899 23.44415 23.70076 23.94007 24.16326 24.37141 24.56553
#> 7  19.70476 19.70476 19.70476 19.70476 19.70476 19.70476 19.70476 19.70476
#> 8  19.70476 19.70476 19.70476 19.70476 19.70476 19.70476 19.70476 19.70476
#> 9  16.53558 16.24054 15.96538 15.70877 15.46945 15.24626 15.03811 14.84400
#> 10 16.53558 16.24054 15.96538 15.70877 15.46945 15.24626 15.03811 14.84400
#> 11 16.53558 16.24054 15.96538 15.70877 15.46945 15.24626 15.03811 14.84400
#> 12 16.53558 16.24054 15.96538 15.70877 15.46945 15.24626 15.03811 14.84400
#> 13 16.53558 16.24054 15.96538 15.70877 15.46945 15.24626 15.03811 14.84400
#> 14 22.87395 23.16899 23.44415 23.70076 23.94007 24.16326 24.37141 24.56553
#> 15 22.87395 23.16899 23.44415 23.70076 23.94007 24.16326 24.37141 24.56553
#> 16 16.53558 16.24054 15.96538 15.70877 15.46945 15.24626 15.03811 14.84400
#> 17 22.87395 23.16899 23.44415 23.70076 23.94007 24.16326 24.37141 24.56553
#> 18 22.87395 23.16899 23.44415 23.70076 23.94007 24.16326 24.37141 24.56553
#> 19 16.53558 16.24054 15.96538 15.70877 15.46945 15.24626 15.03811 14.84400
#> 20 16.53558 16.24054 15.96538 15.70877 15.46945 15.24626 15.03811 14.84400
#> 21 22.87395 23.16899 23.44415 23.70076 23.94007 24.16326 24.37141 24.56553
#>      1.4577   1.3595   1.2679   1.1824   1.1027   1.0284   0.9591   0.8945
#> 1  19.70476 19.70476 19.70476 19.70476 19.70476 19.70476 19.76377 19.89765
#> 2  19.70476 19.70476 19.70476 19.70476 19.70476 19.70476 19.76377 19.89765
#> 3  24.73601 24.73601 24.73601 24.73601 24.73601 24.73601 24.77566 24.86561
#> 4  19.70476 19.70476 19.70476 19.70476 19.70476 19.70476 19.66050 19.56010
#> 5  24.73601 24.73601 24.73601 24.73601 24.73601 24.73601 24.67239 24.52805
#> 6  24.73601 24.73601 24.73601 24.73601 24.73601 24.73601 24.67239 24.52805
#> 7  19.70476 19.70476 19.70476 19.70476 19.70476 19.70476 19.66050 19.56010
#> 8  19.70476 19.70476 19.70476 19.70476 19.70476 19.70476 19.66050 19.56010
#> 9  14.67351 14.67351 14.67351 14.67351 14.67351 14.67351 14.64862 14.59214
#> 10 14.67351 14.67351 14.67351 14.67351 14.67351 14.67351 14.64862 14.59214
#> 11 14.67351 14.67351 14.67351 14.67351 14.67351 14.67351 14.64862 14.59214
#> 12 14.67351 14.67351 14.67351 14.67351 14.67351 14.67351 14.64862 14.59214
#> 13 14.67351 14.67351 14.67351 14.67351 14.67351 14.67351 14.64862 14.59214
#> 14 24.73601 24.73601 24.73601 24.73601 24.73601 24.73601 24.77566 24.86561
#> 15 24.73601 24.73601 24.73601 24.73601 24.73601 24.73601 24.67239 24.52805
#> 16 14.67351 14.67351 14.67351 14.67351 14.67351 14.67351 14.64862 14.59214
#> 17 24.73601 24.73601 24.73601 24.73601 24.73601 24.73601 24.77566 24.86561
#> 18 24.73601 24.73601 24.73601 24.73601 24.73601 24.73601 24.77566 24.86561
#> 19 14.67351 14.67351 14.67351 14.67351 14.67351 14.67351 14.75188 14.92969
#> 20 14.67351 14.67351 14.67351 14.67351 14.67351 14.67351 14.75188 14.92969
#> 21 24.73601 24.73601 24.73601 24.73601 24.73601 24.73601 24.77566 24.86561
#>      0.8342   0.7780   0.7255   0.6766   0.6310   0.5885   0.5488   0.5118
#> 1  20.02256 20.13906 20.24770 20.34902 20.44351 20.53164 20.61382 21.20613
#> 2  20.02256 20.13906 20.24770 20.34902 20.44351 20.53164 20.61382 20.62872
#> 3  24.94953 25.02780 25.10080 25.16887 25.23236 25.29157 25.34678 25.11272
#> 4  19.46641 19.37904 19.29756 19.22157 19.15070 19.08461 19.02297 19.30408
#> 5  24.39338 24.26779 24.15066 24.04142 23.93955 23.84454 23.75593 23.14274
#> 6  24.39338 24.26779 24.15066 24.04142 23.93955 23.84454 23.75593 23.23331
#> 7  19.46641 19.37904 19.29756 19.22157 19.15070 19.08461 19.02297 19.34937
#> 8  19.46641 19.37904 19.29756 19.22157 19.15070 19.08461 19.02297 19.34937
#> 9  14.53944 14.49029 14.44446 14.40172 14.36185 14.32467 14.29000 14.69554
#> 10 14.53944 14.49029 14.44446 14.40172 14.36185 14.32467 14.29000 15.46542
#> 11 14.53944 14.49029 14.44446 14.40172 14.36185 14.32467 14.29000 15.35220
#> 12 14.53944 14.49029 14.44446 14.40172 14.36185 14.32467 14.29000 11.62961
#> 13 14.53944 14.49029 14.44446 14.40172 14.36185 14.32467 14.29000 11.80850
#> 14 24.94953 25.02780 25.10080 25.16887 25.23236 25.29157 25.34678 25.38444
#> 15 24.39338 24.26779 24.15066 24.04142 23.93955 23.84454 23.75593 24.78439
#> 16 14.53944 14.49029 14.44446 14.40172 14.36185 14.32467 14.29000 15.21634
#> 17 24.94953 25.02780 25.10080 25.16887 25.23236 25.29157 25.34678 25.98450
#> 18 24.94953 25.02780 25.10080 25.16887 25.23236 25.29157 25.34678 25.52030
#> 19 15.09559 15.25031 15.39460 15.52917 15.65466 15.77170 15.88085 16.73346
#> 20 15.09559 15.25031 15.39460 15.52917 15.65466 15.77170 15.88085 15.82772
#> 21 24.94953 25.02780 25.10080 25.16887 25.23236 25.29157 25.34678 24.07112
#>      0.4773   0.4452   0.4152   0.3872   0.3611   0.3368   0.3141   0.2929
#> 1  21.20628 21.20637 21.20642 21.20645 21.20646 21.20647 22.62386 22.64108
#> 2  20.62881 20.62887 20.62890 20.62892 20.62893 20.62893 22.25356 22.27632
#> 3  25.11264 25.11259 25.11256 25.11254 25.11253 25.11252 25.23715 25.24732
#> 4  19.30404 19.30402 19.30400 19.30400 19.30399 19.30399 19.68062 19.67315
#> 5  23.14246 23.14230 23.14220 23.14215 23.14211 23.14209 22.89015 22.88097
#> 6  23.23305 23.23289 23.23279 23.23274 23.23271 23.23269 22.13617 22.12019
#> 7  19.34933 19.34931 19.34930 19.34929 19.34929 19.34929 19.26671 19.25558
#> 8  19.34933 19.34931 19.34930 19.34929 19.34929 19.34929 19.26671 19.25558
#> 9  14.69567 14.69574 14.69578 14.69581 14.69582 14.69583 15.18989 15.18579
#> 10 15.46562 15.46574 15.46581 15.46585 15.46587 15.46588 15.68363 15.67214
#> 11 15.35239 15.35250 15.35257 15.35261 15.35263 15.35264 15.61102 15.60062
#> 12 11.62944 11.62934 11.62928 11.62924 11.62922 11.62921 12.36239 12.38142
#> 13 11.80834 11.80825 11.80819 11.80816 11.80814 11.80813 12.10799 12.12260
#> 14 25.38439 25.38435 25.38433 25.38432 25.38431 25.38431 26.07583 26.08824
#> 15 24.78427 24.78421 24.78417 24.78414 24.78413 24.78412 23.08168 23.05046
#> 16 15.21652 15.21662 15.21668 15.21672 15.21674 15.21675 13.92437 13.90359
#> 17 25.98450 25.98450 25.98450 25.98450 25.98450 25.98450 26.46065 26.46731
#> 18 25.52026 25.52023 25.52022 25.52021 25.52020 25.52020 25.54775 25.55438
#> 19 16.73378 16.73397 16.73408 16.73415 16.73419 16.73421 16.27624 16.28132
#> 20 15.82795 15.82809 15.82817 15.82822 15.82825 15.82826 13.94820 13.94922
#> 21 24.07094 24.07083 24.07076 24.07072 24.07070 24.07069 24.17543 24.19272
#>      0.2732   0.2547   0.2376   0.2216   0.2066   0.1927   0.1797   0.1676
#> 1  22.64142 22.64170 22.64193 22.64212 22.64227 22.64239 22.64248 22.64256
#> 2  22.27669 22.27699 22.27723 22.27742 22.27757 22.27769 22.27778 22.27785
#> 3  25.24733 25.24733 25.24731 25.24729 25.24727 25.24724 25.24722 25.24721
#> 4  19.67323 19.67330 19.67335 19.67340 19.67344 19.67347 19.67349 19.67351
#> 5  22.88086 22.88075 22.88065 22.88056 22.88048 22.88042 22.88036 22.88031
#> 6  22.11990 22.11964 22.11943 22.11925 22.11910 22.11898 22.11888 22.11880
#> 7  19.25556 19.25555 19.25554 19.25554 19.25554 19.25554 19.25554 19.25554
#> 8  19.25556 19.25555 19.25554 19.25554 19.25554 19.25554 19.25554 19.25554
#> 9  15.18592 15.18605 15.18616 15.18626 15.18634 15.18641 15.18647 15.18652
#> 10 15.67222 15.67232 15.67242 15.67252 15.67260 15.67268 15.67274 15.67280
#> 11 15.60071 15.60081 15.60091 15.60101 15.60109 15.60117 15.60123 15.60129
#> 12 12.38154 12.38162 12.38166 12.38167 12.38168 12.38167 12.38166 12.38165
#> 13 12.12264 12.12264 12.12263 12.12260 12.12257 12.12254 12.12251 12.12249
#> 14 26.08838 26.08849 26.08856 26.08861 26.08865 26.08867 26.08869 26.08870
#> 15 23.05006 23.04975 23.04950 23.04931 23.04916 23.04904 23.04895 23.04888
#> 16 13.90334 13.90317 13.90304 13.90296 13.90291 13.90287 13.90285 13.90283
#> 17 26.46741 26.46749 26.46756 26.46761 26.46764 26.46767 26.46769 26.46771
#> 18 25.55437 25.55436 25.55435 25.55433 25.55431 25.55429 25.55428 25.55427
#> 19 16.28128 16.28127 16.28128 16.28130 16.28132 16.28135 16.28137 16.28140
#> 20 13.94885 13.94856 13.94834 13.94817 13.94804 13.94794 13.94786 13.94780
#> 21 24.19271 24.19267 24.19261 24.19255 24.19249 24.19244 24.19239 24.19234
#>      0.1563   0.1458   0.1359   0.1268   0.1182   0.1103   0.1028   0.0959
#> 1  22.64262 22.64267 22.46040 22.62741 22.62734 22.21880 22.21921 22.21957
#> 2  22.27790 22.27795 22.29129 22.51220 22.51213 22.11616 22.11652 22.11685
#> 3  25.24719 25.24718 25.11866 24.85143 24.85146 25.04610 25.04589 25.04571
#> 4  19.67353 19.67355 19.98794 19.69448 19.69430 19.91938 19.91910 19.91888
#> 5  22.88027 22.88024 22.67582 22.73551 22.73567 22.75492 22.75537 22.75573
#> 6  22.11873 22.11868 23.17478 23.69339 23.69348 23.63548 23.63528 23.63512
#> 7  19.25554 19.25555 19.21997 18.83853 18.83841 18.63887 18.63931 18.63970
#> 8  19.25554 19.25555 19.47593 19.18046 19.18031 18.95340 18.95375 18.95407
#> 9  15.18656 15.18660 15.13799 15.21973 15.21965 15.33494 15.33484 15.33477
#> 10 15.67285 15.67289 15.76731 15.91284 15.91271 15.96806 15.96788 15.96775
#> 11 15.60133 15.60137 15.85795 16.05562 16.05548 16.10005 16.09982 16.09965
#> 12 12.38164 12.38162 12.29995 12.34364 12.34374 12.39776 12.39783 12.39788
#> 13 12.12246 12.12244 11.89110 11.90348 11.90362 11.99182 11.99185 11.99186
#> 14 26.08871 26.08871 26.33385 26.17013 26.17004 26.25758 26.25739 26.25723
#> 15 23.04883 23.04879 22.53268 22.54667 22.54691 22.44615 22.44617 22.44617
#> 16 13.90282 13.90282 13.07705 12.97473 12.97486 12.95755 12.95756 12.95754
#> 17 26.46772 26.46773 26.51470 26.29667 26.29659 26.37053 26.37038 26.37026
#> 18 25.55426 25.55425 25.09712 25.73345 25.73372 26.03904 26.03925 26.03938
#> 19 16.28142 16.28144 16.60963 16.54154 16.54137 17.14634 17.14567 17.14511
#> 20 13.94776 13.94773 14.28642 14.27578 14.27581 13.83043 13.83028 13.83017
#> 21 24.19231 24.19227 23.98946 23.69231 23.69242 23.67664 23.67664 23.67662
#>      0.0894   0.0834   0.0778   0.0726   0.0677   0.0631   0.0588   0.0549
#> 1  22.21990 22.22018 22.22043 22.22064 22.22082 21.80220 21.80191 21.80166
#> 2  22.11714 22.11740 22.11763 22.11783 22.11800 22.08286 22.08258 22.08233
#> 3  25.04554 25.04540 25.04527 25.04516 25.04507 25.10783 25.10806 25.10827
#> 4  19.91870 19.91858 19.91848 19.91842 19.91838 19.18976 19.19016 19.19051
#> 5  22.75601 22.75622 22.75638 22.75650 22.75658 22.31502 22.31485 22.31472
#> 6  23.63498 23.63486 23.63476 23.63468 23.63462 23.80878 23.80860 23.80843
#> 7  18.64004 18.64034 18.64059 18.64080 18.64099 18.88831 18.88815 18.88801
#> 8  18.95435 18.95461 18.95483 18.95503 18.95520 19.18902 19.18887 19.18873
#> 9  15.33472 15.33469 15.33466 15.33465 15.33465 16.07052 16.07050 16.07048
#> 10 15.96765 15.96758 15.96754 15.96751 15.96751 16.17075 16.17073 16.17072
#> 11 16.09952 16.09943 16.09937 16.09933 16.09932 16.37122 16.37121 16.37120
#> 12 12.39791 12.39793 12.39793 12.39792 12.39791 11.78202 11.78227 11.78249
#> 13 11.99184 11.99182 11.99178 11.99173 11.99168 12.02725 12.02741 12.02755
#> 14 26.25711 26.25701 26.25694 26.25689 26.25686 26.49261 26.49279 26.49293
#> 15 22.44614 22.44610 22.44604 22.44597 22.44590 22.29688 22.29652 22.29622
#> 16 12.95751 12.95747 12.95742 12.95736 12.95730 12.99243 12.99227 12.99215
#> 17 26.37016 26.37009 26.37004 26.37000 26.36998 26.20137 26.20154 26.20168
#> 18 26.03945 26.03948 26.03947 26.03944 26.03938 26.14200 26.14182 26.14168
#> 19 17.14464 17.14426 17.14395 17.14370 17.14351 16.89856 16.89929 16.89988
#> 20 13.83008 13.83002 13.82997 13.82994 13.82992 13.76019 13.76002 13.75988
#> 21 23.67659 23.67655 23.67651 23.67647 23.67642 24.21041 24.21045 24.21050
#>      0.0512   0.0477   0.0445   0.0415   0.0387   0.0361   0.0337   0.0314
#> 1  21.80144 21.80125 21.80110 21.80096 21.80085 21.80076 21.80068 21.80062
#> 2  22.08211 22.08191 22.08174 22.08159 22.08145 22.08134 22.08124 22.08116
#> 3  25.10845 25.10862 25.10877 25.10890 25.10902 25.10912 25.10920 25.10928
#> 4  19.19081 19.19108 19.19130 19.19150 19.19166 19.19180 19.19191 19.19201
#> 5  22.31464 22.31458 22.31455 22.31454 22.31455 22.31458 22.31461 22.31465
#> 6  23.80826 23.80810 23.80795 23.80781 23.80768 23.80757 23.80746 23.80736
#> 7  18.88788 18.88776 18.88765 18.88755 18.88747 18.88739 18.88732 18.88726
#> 8  19.18859 19.18846 19.18834 19.18822 19.18811 19.18801 19.18792 19.18784
#> 9  16.07046 16.07045 16.07043 16.07042 16.07041 16.07040 16.07039 16.07039
#> 10 16.17070 16.17068 16.17066 16.17064 16.17063 16.17061 16.17059 16.17058
#> 11 16.37117 16.37115 16.37112 16.37109 16.37106 16.37103 16.37100 16.37097
#> 12 11.78267 11.78283 11.78297 11.78308 11.78317 11.78325 11.78332 11.78337
#> 13 12.02767 12.02777 12.02785 12.02792 12.02798 12.02802 12.02806 12.02809
#> 14 26.49304 26.49312 26.49318 26.49321 26.49324 26.49324 26.49324 26.49323
#> 15 22.29600 22.29582 22.29570 22.29561 22.29556 22.29553 22.29553 22.29554
#> 16 12.99205 12.99197 12.99191 12.99186 12.99183 12.99181 12.99180 12.99179
#> 17 26.20180 26.20189 26.20196 26.20202 26.20206 26.20209 26.20211 26.20212
#> 18 26.14158 26.14152 26.14149 26.14148 26.14149 26.14152 26.14155 26.14160
#> 19 16.90038 16.90077 16.90109 16.90133 16.90152 16.90166 16.90175 16.90181
#> 20 13.75978 13.75970 13.75965 13.75962 13.75961 13.75961 13.75962 13.75964
#> 21 24.21053 24.21057 24.21060 24.21063 24.21065 24.21067 24.21069 24.21071
#>      0.0293   0.0273   0.0255   0.0238   0.0222   0.0207   0.0193   0.0180
#> 1  21.80057 21.80053 21.58605 21.58594 21.58583 21.58573 21.58563 21.58553
#> 2  22.08109 22.08103 22.03166 22.03164 22.03162 22.03161 22.03159 22.03158
#> 3  25.10934 25.10940 25.20820 25.20823 25.20827 25.20830 25.20833 25.20837
#> 4  19.19209 19.19215 19.00787 19.00777 19.00768 19.00759 19.00750 19.00741
#> 5  22.31469 22.31474 22.29057 22.29052 22.29048 22.29044 22.29040 22.29036
#> 6  23.80728 23.80720 23.74239 23.74238 23.74236 23.74235 23.74233 23.74232
#> 7  18.88721 18.88716 18.97997 18.98007 18.98016 18.98024 18.98033 18.98041
#> 8  19.18776 19.18770 19.23489 19.23497 19.23504 19.23511 19.23519 19.23525
#> 9  16.07038 16.07038 16.36483 16.36502 16.36520 16.36538 16.36555 16.36572
#> 10 16.17057 16.17055 16.17289 16.17293 16.17297 16.17300 16.17304 16.17307
#> 11 16.37094 16.37091 16.38356 16.38361 16.38365 16.38370 16.38375 16.38379
#> 12 11.78340 11.78343 11.74994 11.74983 11.74974 11.74964 11.74954 11.74945
#> 13 12.02811 12.02812 12.19747 12.19749 12.19751 12.19753 12.19754 12.19756
#> 14 26.49321 26.49319 26.54992 26.54998 26.55004 26.55009 26.55014 26.55020
#> 15 22.29557 22.29560 22.10985 22.10972 22.10959 22.10947 22.10935 22.10924
#> 16 12.99179 12.99180 12.86596 12.86588 12.86580 12.86572 12.86565 12.86557
#> 17 26.20212 26.20212 26.08417 26.08413 26.08409 26.08405 26.08401 26.08398
#> 18 26.14165 26.14170 26.25440 26.25445 26.25449 26.25454 26.25458 26.25462
#> 19 16.90184 16.90185 16.75433 16.75430 16.75427 16.75425 16.75422 16.75419
#> 20 13.75966 13.75969 13.74085 13.74079 13.74074 13.74068 13.74063 13.74058
#> 21 24.21072 24.21074 24.49024 24.49036 24.49047 24.49058 24.49069 24.49079
#>      0.0168   0.0156   0.0146   0.0136   0.0127   0.0118   0.0110   0.0103
#> 1  21.58544 21.58534 21.58525 21.58517 21.58508 21.58500 21.58492 21.58484
#> 2  22.03156 22.03155 22.03154 22.03152 22.03151 22.03150 22.03149 22.03147
#> 3  25.20840 25.20843 25.20846 25.20849 25.20851 25.20854 25.20857 25.20859
#> 4  19.00732 19.00724 19.00716 19.00708 19.00700 19.00693 19.00686 19.00679
#> 5  22.29032 22.29028 22.29024 22.29021 22.29017 22.29014 22.29011 22.29008
#> 6  23.74231 23.74230 23.74228 23.74227 23.74226 23.74225 23.74224 23.74223
#> 7  18.98049 18.98057 18.98065 18.98072 18.98079 18.98086 18.98093 18.98100
#> 8  19.23532 19.23538 19.23545 19.23551 19.23557 19.23563 19.23568 19.23574
#> 9  16.36588 16.36604 16.36620 16.36635 16.36649 16.36663 16.36677 16.36691
#> 10 16.17311 16.17314 16.17317 16.17320 16.17323 16.17326 16.17329 16.17332
#> 11 16.38383 16.38387 16.38391 16.38395 16.38399 16.38403 16.38407 16.38410
#> 12 11.74936 11.74928 11.74919 11.74911 11.74903 11.74896 11.74888 11.74881
#> 13 12.19758 12.19760 12.19762 12.19763 12.19765 12.19766 12.19768 12.19769
#> 14 26.55024 26.55029 26.55034 26.55039 26.55043 26.55047 26.55052 26.55056
#> 15 22.10913 22.10902 22.10891 22.10881 22.10871 22.10861 22.10852 22.10843
#> 16 12.86550 12.86543 12.86536 12.86530 12.86523 12.86517 12.86511 12.86505
#> 17 26.08394 26.08391 26.08387 26.08384 26.08381 26.08378 26.08375 26.08372
#> 18 26.25466 26.25470 26.25474 26.25478 26.25482 26.25485 26.25489 26.25492
#> 19 16.75417 16.75414 16.75412 16.75410 16.75408 16.75405 16.75403 16.75401
#> 20 13.74053 13.74048 13.74044 13.74039 13.74035 13.74030 13.74026 13.74022
#> 21 24.49090 24.49100 24.49109 24.49119 24.49128 24.49137 24.49145 24.49154
#>      0.0096   0.0089   0.0083   0.0078   0.0073   0.0068   0.0063   0.0059
#> 1  21.58476 21.58469 21.58462 21.58455 21.58448 21.58441 21.58435 21.58429
#> 2  22.03146 22.03145 22.03144 22.03143 22.03142 22.03141 22.03140 22.03139
#> 3  25.20862 25.20864 25.20867 25.20869 25.20871 25.20873 25.20875 25.20877
#> 4  19.00672 19.00665 19.00659 19.00653 19.00646 19.00641 19.00635 19.00629
#> 5  22.29004 22.29001 22.28999 22.28996 22.28993 22.28990 22.28988 22.28985
#> 6  23.74222 23.74221 23.74220 23.74219 23.74218 23.74217 23.74216 23.74215
#> 7  18.98106 18.98112 18.98118 18.98124 18.98130 18.98136 18.98141 18.98147
#> 8  19.23579 19.23584 19.23589 19.23594 19.23599 19.23604 19.23608 19.23613
#> 9  16.36704 16.36716 16.36729 16.36741 16.36752 16.36764 16.36775 16.36786
#> 10 16.17334 16.17337 16.17340 16.17342 16.17345 16.17347 16.17349 16.17351
#> 11 16.38414 16.38417 16.38420 16.38423 16.38426 16.38429 16.38432 16.38435
#> 12 11.74874 11.74867 11.74860 11.74853 11.74847 11.74841 11.74835 11.74829
#> 13 12.19771 12.19772 12.19774 12.19775 12.19776 12.19777 12.19779 12.19780
#> 14 26.55060 26.55064 26.55067 26.55071 26.55075 26.55078 26.55081 26.55085
#> 15 22.10834 22.10825 22.10816 22.10808 22.10800 22.10792 22.10785 22.10777
#> 16 12.86500 12.86494 12.86489 12.86483 12.86478 12.86473 12.86468 12.86464
#> 17 26.08369 26.08366 26.08363 26.08361 26.08358 26.08356 26.08353 26.08351
#> 18 26.25495 26.25498 26.25501 26.25504 26.25507 26.25510 26.25513 26.25516
#> 19 16.75399 16.75397 16.75395 16.75393 16.75392 16.75390 16.75388 16.75387
#> 20 13.74018 13.74014 13.74011 13.74007 13.74003 13.74000 13.73997 13.73993
#> 21 24.49162 24.49170 24.49178 24.49185 24.49192 24.49199 24.49206 24.49213
#>      0.0055   0.0051   0.0048   0.0045
#> 1  21.58423 21.58417 21.58411 21.58405
#> 2  22.03138 22.03137 22.03136 22.03136
#> 3  25.20879 25.20881 25.20883 25.20885
#> 4  19.00624 19.00618 19.00613 19.00608
#> 5  22.28983 22.28980 22.28978 22.28976
#> 6  23.74214 23.74214 23.74213 23.74212
#> 7  18.98152 18.98157 18.98162 18.98166
#> 8  19.23617 19.23621 19.23625 19.23629
#> 9  16.36796 16.36806 16.36816 16.36826
#> 10 16.17354 16.17356 16.17358 16.17360
#> 11 16.38438 16.38440 16.38443 16.38446
#> 12 11.74823 11.74818 11.74812 11.74807
#> 13 12.19781 12.19782 12.19783 12.19784
#> 14 26.55088 26.55091 26.55094 26.55097
#> 15 22.10770 22.10763 22.10756 22.10750
#> 16 12.86459 12.86455 12.86450 12.86446
#> 17 26.08349 26.08347 26.08344 26.08342
#> 18 26.25518 26.25521 26.25523 26.25526
#> 19 16.75385 16.75383 16.75382 16.75380
#> 20 13.73990 13.73987 13.73984 13.73981
#> 21 24.49220 24.49226 24.49232 24.49238
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 22.8 18.1 24.4 22.8 19.2 17.8 16.4 17.3 15.2 10.4 14.7 32.4 21.5
#> [16] 13.3 27.3 26.0 15.8 15.0 21.4
#> 
#> $X
#>               am        carb       cyl        disp        drat       gear
#>  [1,]  1.1547005  0.63614643  0.000000 -0.49632657  0.49523118  0.3483665
#>  [2,]  1.1547005  0.63614643  0.000000 -0.49632657  0.49523118  0.3483665
#>  [3,]  1.1547005 -1.18554561 -1.145644 -0.96187254  0.38482932  0.3483665
#>  [4,] -0.8660254 -1.18554561  0.000000  0.08560589 -2.02193112 -1.1147728
#>  [5,] -0.8660254 -0.57831493 -1.145644 -0.61539891  0.03154339  0.3483665
#>  [6,] -0.8660254 -0.57831493 -1.145644 -0.66822047  0.53939192  0.3483665
#>  [7,] -0.8660254  0.63614643  0.000000 -0.42828524  0.53939192  0.3483665
#>  [8,] -0.8660254  0.63614643  0.000000 -0.42828524  0.53939192  0.3483665
#>  [9,] -0.8660254  0.02891575  1.145644  0.54040849 -1.33743962 -1.1147728
#> [10,] -0.8660254  0.02891575  1.145644  0.54040849 -1.33743962 -1.1147728
#> [11,] -0.8660254  0.02891575  1.145644  0.54040849 -1.33743962 -1.1147728
#> [12,] -0.8660254  0.63614643  1.145644  2.18951555 -1.49200221 -1.1147728
#> [13,] -0.8660254  0.63614643  1.145644  2.01045941 -0.98415368 -1.1147728
#> [14,]  1.1547005 -1.18554561 -1.145644 -1.22418979  0.89267786  0.3483665
#> [15,] -0.8660254 -1.18554561 -1.145644 -0.85354357  0.05362376 -1.1147728
#> [16,] -0.8660254  0.63614643  1.145644  1.20470677  0.11986487 -1.1147728
#> [17,]  1.1547005 -1.18554561 -1.145644 -1.22150394  0.89267786  0.3483665
#> [18,]  1.1547005 -0.57831493 -1.145644 -0.85175301  1.66549084  1.8115058
#> [19,]  1.1547005  0.63614643  1.145644  1.21365958  1.20180305  1.8115058
#> [20,]  1.1547005  3.06506914  1.145644  0.76601923 -0.29966218  1.8115058
#> [21,]  1.1547005 -0.57831493 -1.145644 -0.84548605  0.95891897  0.3483665
#>               hp        qsec         vs         wt
#>  [1,] -0.5011698 -0.81994369 -0.9534626 -0.7333592
#>  [2,] -0.5011698 -0.52933074 -0.9534626 -0.4513182
#>  [3,] -0.7338326  0.29580247  1.0488088 -1.0651722
#>  [4,] -0.5696000  1.13131472  1.0488088  0.1957170
#>  [5,] -1.1580999  1.01714534  1.0488088 -0.1029147
#>  [6,] -0.7064605  2.52210529  1.0488088 -0.1471564
#>  [7,] -0.3232513  0.13492744  1.0488088  0.1735961
#>  [8,] -0.3232513  0.44629847  1.0488088  0.1735961
#>  [9,]  0.4568531 -0.33212909 -0.9534626  0.8704032
#> [10,]  0.4568531 -0.22833875 -0.9534626  0.4943486
#> [11,]  0.4568531 -0.02075807 -0.9534626  0.5496507
#> [12,]  0.9358646 -0.11416938 -0.9534626  2.3679856
#> [13,]  1.1411552 -0.32175006 -0.9534626  2.2806082
#> [14,] -1.1033557  0.74210094  1.0488088 -1.1978973
#> [15,] -0.6790884  1.02233486  1.0488088 -0.9047959
#> [16,]  1.3464459 -1.36484298 -0.9534626  0.6160133
#> [17,] -1.1033557  0.44629847  1.0488088 -1.4909987
#> [18,] -0.7612046 -0.69539528 -0.9534626 -1.2642599
#> [19,]  1.6064807 -1.83708904 -0.9534626 -0.1250355
#> [20,]  2.5781897 -1.78519386 -0.9534626  0.3173817
#> [21,] -0.5148559  0.29061295  1.0488088 -0.5563923
#> attr(,"center")
#>  [1]   0.4285714   2.9523810   6.0000000 215.4380952   3.6757143   3.7619048
#>  [7] 146.6190476  18.0400000   0.4761905   3.2830476
#> attr(,"scale")
#>  [1]   0.4948717   1.6468206   1.7457431 111.6968110   0.4528909   0.6834619
#>  [7]  73.0671421   1.9269616   0.4994328   0.9041239
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 37
#> 
#> $lambda.min
#> [1] 0.3610932
#> 
#> $null.dev
#> [1] 25.8395
#> 
#> $Bias
#> [1] 4.94733
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 10.64418