Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157. doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

LearnerRegrNCVreg$new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

LearnerRegrNCVreg$clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 41.32345 38.02835 34.30971 30.59746 27.55161 25.07306 23.07640 21.48816
#>   [9] 20.24513 19.29303 18.58528 18.08192 17.74878 17.55664 17.48055 17.49818
#>  [17] 17.16741 16.84175 16.73569 16.64123 16.44224 16.24973 15.94513 15.68050
#>  [25] 15.45059 15.23347 12.91020 11.94225 11.34879 11.11744 10.93959 10.81780
#>  [33] 10.74258 15.33673 14.91666 14.46107 14.17672 14.15748 14.14163 14.56453
#>  [41] 14.42790 14.63596 15.08566 15.05664 16.11049 16.91850 16.96345 16.68985
#>  [49] 15.69121 15.96987 15.97940 16.86762 20.31521 21.83518 22.12357 22.42470
#>  [57] 22.97156 23.00672 22.92475 22.95669 23.12724 22.58744 22.58709 24.86230
#>  [65] 25.24569 25.25038 25.50498 25.50417 25.50336 25.50271 25.55813 25.55735
#>  [73] 25.55659 25.55592 25.09321 25.05611 25.05548 25.05488 25.06880 25.01641
#>  [81] 25.01598 24.92027 24.88658 24.88720 24.70466 25.30381 25.30433 25.51043
#>  [89] 25.51154 25.51260 25.51363 25.51462 25.51558 25.51651 25.48725 25.49364
#>  [97] 25.49427 25.46287 25.46330 25.49940
#> 
#> $cvse
#>   [1] 10.753178 10.475449  9.730336  8.825400  8.003502  7.256845  6.580531
#>   [8]  5.972097  5.431136  4.958898  4.557799  4.230696  3.979863  3.805776
#>  [15]  3.706021  3.674329  3.578821  3.473257  3.445747  3.397991  3.306940
#>  [22]  3.227838  3.185140  3.157465  3.132873  3.095394  2.854196  2.792943
#>  [29]  2.756544  2.677163  2.614394  2.570574  2.537798  5.415824  5.429949
#>  [36]  5.390648  5.377166  5.368479  5.360877  5.365066  5.356790  5.304121
#>  [43]  5.273071  5.271382  5.272539  5.426303  5.431613  5.468431  4.574623
#>  [50]  4.582198  4.581818  5.042604  8.098589  8.296704  8.326457  8.316091
#>  [57]  8.290132  8.285296  8.285775  8.298520  8.302538  8.230746  8.230618
#>  [64]  9.116877  9.010798  9.009525  8.993062  8.992213  8.991396  8.990595
#>  [71]  8.989970  8.989244  8.988546  8.987866  9.070313  9.079681  9.079117
#>  [78]  9.078576  9.079971  8.989146  8.988728  8.978587  8.982320  8.981933
#>  [85]  8.978254  9.517384  9.518637  9.508730  9.509939  9.511125  9.512288
#>  [92]  9.513427  9.514545  9.515640  9.517376  9.519073  9.520117  9.515603
#>  [99]  9.516585  9.521979
#> 
#> $fold
#>  [1]  8  9  5  2  3  7  1  6  8  2 10  4  1  5 10  9  6  7  3  4  1
#> 
#> $lambda
#>   [1] 5.485186320 5.115503120 4.770735331 4.449203736 4.149342296 3.869690512
#>   [7] 3.608886323 3.365659463 3.138825280 2.927278962 2.729990157 2.545997957
#>  [13] 2.374406216 2.214379184 2.065137438 1.925954087 1.796151227 1.675096646
#>  [19] 1.562200738 1.456913637 1.358722534 1.267149183 1.181747569 1.102101738
#>  [25] 1.027823769 0.958551887 0.893948698 0.833699548 0.777510989 0.725109350
#>  [31] 0.676239407 0.630663134 0.588158550 0.548518632 0.511550312 0.477073533
#>  [37] 0.444920374 0.414934230 0.386969051 0.360888632 0.336565946 0.313882528
#>  [43] 0.292727896 0.272999016 0.254599796 0.237440622 0.221437918 0.206513744
#>  [49] 0.192595409 0.179615123 0.167509665 0.156220074 0.145691364 0.135872253
#>  [55] 0.126714918 0.118174757 0.110210174 0.102782377 0.095855189 0.089394870
#>  [61] 0.083369955 0.077751099 0.072510935 0.067623941 0.063066313 0.058815855
#>  [67] 0.054851863 0.051155031 0.047707353 0.044492037 0.041493423 0.038696905
#>  [73] 0.036088863 0.033656595 0.031388253 0.029272790 0.027299902 0.025459980
#>  [79] 0.023744062 0.022143792 0.020651374 0.019259541 0.017961512 0.016750966
#>  [85] 0.015622007 0.014569136 0.013587225 0.012671492 0.011817476 0.011021017
#>  [91] 0.010278238 0.009585519 0.008939487 0.008336995 0.007775110 0.007251094
#>  [97] 0.006762394 0.006306631 0.005881585 0.005485186
#> 
#> $fit
#> $beta
#>               5.4852     5.1155     4.7707     4.4492     4.1493    3.8697
#> (Intercept) 19.76667 21.1569587 22.8853002 24.4971574 26.0003807 27.402292
#> am           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> cyl          0.00000 -0.2211828 -0.4961462 -0.7525781 -0.9917272 -1.214759
#> disp         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> wt           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#>                3.6089   3.3657    3.1388    2.9273    2.7300    2.5460
#> (Intercept) 28.709719 29.92903 31.066163 32.126657 33.115678 34.038042
#> am           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.422758 -1.61674 -1.797647 -1.966362 -2.123706 -2.270446
#> disp         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#>                2.3744    2.2144    2.0651    1.9260    1.7962    1.6751
#> (Intercept) 34.898242 35.700467 36.448624 37.146359 37.789744 37.789744
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.407296 -2.534923 -2.653948 -2.764951 -2.867308 -2.867308
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.5622    1.4569    1.3587    1.2671    1.1817    1.1021
#> (Intercept) 37.789744 37.789744 37.789744 37.789744 37.789744 38.351721
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.867308 -2.867308 -2.867308 -2.867308 -2.867308 -2.209640
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000 -1.429972
#>                 1.0278    0.9586    0.8939   0.8337    0.7775    0.7251
#> (Intercept) 37.7984044 38.901208 38.901468 38.90163 38.901732 38.901795
#> am           0.0000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> carb         0.0000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> cyl         -0.9594914 -1.561149 -1.561291 -1.56138 -1.561435 -1.561469
#> disp         0.0000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> drat         0.0000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> gear         0.0000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> hp           0.0000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> qsec         0.0000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> vs           0.0000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> wt          -3.6543840 -2.838573 -2.838381 -2.83826 -2.838186 -2.838139
#>                0.6762    0.6307    0.5882      0.5485     0.5116    0.4771
#> (Intercept) 38.901834 38.901859 38.901874 38.86668196 38.8295564 38.794889
#> am           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.000000
#> carb         0.000000  0.000000  0.000000 -0.06003072 -0.1231193 -0.182003
#> cyl         -1.561491 -1.561504 -1.561513 -1.52777105 -1.4922339 -1.459056
#> disp         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.000000
#> wt          -2.838110 -2.838091 -2.838080 -2.83971897 -2.8415454 -2.843262
#>                0.4449     0.4149     0.3870    0.3609     0.3366    0.3139
#> (Intercept) 38.762556 38.7324033 38.7042824 38.678057 38.6535985 38.630789
#> am           0.000000  0.0000000  0.0000000  0.000000  0.0000000  0.000000
#> carb        -0.236919 -0.2881338 -0.3358969 -0.380441 -0.4219829 -0.460725
#> cyl         -1.428114 -1.3992570 -1.3723450 -1.347247 -1.3238400 -1.302011
#> disp         0.000000  0.0000000  0.0000000  0.000000  0.0000000  0.000000
#> drat         0.000000  0.0000000  0.0000000  0.000000  0.0000000  0.000000
#> gear         0.000000  0.0000000  0.0000000  0.000000  0.0000000  0.000000
#> hp           0.000000  0.0000000  0.0000000  0.000000  0.0000000  0.000000
#> qsec         0.000000  0.0000000  0.0000000  0.000000  0.0000000  0.000000
#> vs           0.000000  0.0000000  0.0000000  0.000000  0.0000000  0.000000
#> wt          -2.844864 -2.8463576 -2.8477506 -2.849050 -2.8502613 -2.851391
#>                 0.2927      0.2730      0.2546      0.2374      0.2214
#> (Intercept) 38.6106869 29.15150287 29.15345111 29.15482470 29.15579314
#> am           0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb        -0.4949825  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl         -1.2827452  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.0000000  1.51093285  1.51052626  1.51023960  1.51003749
#> hp           0.0000000 -0.04388808 -0.04388321 -0.04387978 -0.04387736
#> qsec         0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -2.8523404 -2.40526169 -2.40563841 -2.40590401 -2.40609127
#>                  0.2065      0.1926     0.1796    0.1675      0.1562
#> (Intercept) 29.15647593 29.15695733 29.1572967 29.157536 29.15770473
#> am           0.00000000  0.00000000  0.0000000  0.000000  0.00000000
#> carb         0.00000000  0.00000000  0.0000000  0.000000  0.00000000
#> cyl          0.00000000  0.00000000  0.0000000  0.000000  0.00000000
#> disp         0.00000000  0.00000000  0.0000000  0.000000  0.00000000
#> drat         0.00000000  0.00000000  0.0000000  0.000000  0.00000000
#> gear         1.50989499  1.50979453  1.5097237  1.509674  1.50963854
#> hp          -0.04387565 -0.04387445 -0.0438736 -0.043873 -0.04387258
#> qsec         0.00000000  0.00000000  0.0000000  0.000000  0.00000000
#> vs           0.00000000  0.00000000  0.0000000  0.000000  0.00000000
#> wt          -2.40622330 -2.40631638 -2.4063820 -2.406428 -2.40646090
#>                  0.1457      0.1359     0.1267      0.1182      0.1102
#> (Intercept) 27.48722568 27.48894342 27.4902582 27.49125447 27.49200165
#> am           0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> drat         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> gear         1.39538748  1.39504686  1.3948007  1.39462540  1.39450259
#> hp          -0.03722753 -0.03722344 -0.0372207 -0.03721892 -0.03721782
#> qsec         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> vs           1.41857301  1.41851631  1.4184287  1.41832840  1.41822685
#> wt          -2.27929565 -2.27963040 -2.2798780 -2.28005912 -2.28018981
#>                  0.1028      0.0959       0.0894       0.0834       0.0778
#> (Intercept) 27.49255600 27.49296259 18.080260978 18.077971860 18.075809470
#> am           0.00000000  0.00000000  0.000000000  0.000000000  0.000000000
#> carb         0.00000000  0.00000000  0.000000000  0.000000000  0.000000000
#> cyl          0.00000000  0.00000000  0.888817217  0.889116366  0.889399242
#> disp         0.00000000  0.00000000  0.002519498  0.002519055  0.002518647
#> drat         0.00000000  0.00000000  0.570297615  0.570291570  0.570285118
#> gear         1.39441827  1.39436181  1.869161282  1.869261220  1.869355648
#> hp          -0.03721719 -0.03721687 -0.045725596 -0.045726431 -0.045727213
#> qsec         0.00000000  0.00000000  0.000000000  0.000000000  0.000000000
#> vs           1.41813088  1.41804415  3.510655746  3.511470558  3.512242984
#> wt          -2.28028277 -2.28034778 -2.319467531 -2.319479563 -2.319492242
#>                   0.0725      0.0676       0.0631      0.0588       0.0549
#> (Intercept) 18.073767609 18.07184034 18.070021978 21.82930354 21.831461444
#> am           0.000000000  0.00000000  0.000000000  1.15864591  1.158569729
#> carb         0.000000000  0.00000000  0.000000000  0.32708658  0.327023244
#> cyl          0.889666627  0.88991927  0.890157885  0.93177370  0.931453231
#> disp         0.002518273  0.00251793  0.002517617  0.00641826  0.006418807
#> drat         0.570278317  0.57027122  0.570263870  0.00000000  0.000000000
#> gear         1.869444835  1.86952904  1.869608507  1.22443750  1.224407546
#> hp          -0.045727946 -0.04572863 -0.045729273 -0.05216153 -0.052159104
#> qsec         0.000000000  0.00000000  0.000000000  0.00000000  0.000000000
#> vs           3.512974958  3.51366834  3.514324931  4.03792484  4.037096457
#> wt          -2.319505470 -2.31951916 -2.319533223 -2.69195051 -2.691948027
#>                   0.0512      0.0477       0.0445       0.0415       0.0387
#> (Intercept) 21.833540574 21.83554297 21.837470639 21.839325592 21.841109802
#> am           1.158493260  1.15841664  1.158339986  1.158263417  1.158187039
#> carb         0.326960366  0.32689801  0.326836233  0.326775090  0.326714629
#> cyl          0.931143440  0.93084409  0.930554953  0.930275786  0.930006358
#> disp         0.006419293  0.00641972  0.006420091  0.006420411  0.006420682
#> drat         0.000000000  0.00000000  0.000000000  0.000000000  0.000000000
#> gear         1.224381145  1.22435811  1.224338241  1.224321378  1.224307345
#> hp          -0.052156730 -0.05215441 -0.052152142 -0.052149929 -0.052147769
#> qsec         0.000000000  0.00000000  0.000000000  0.000000000  0.000000000
#> vs           4.036291362  4.03550923  4.034749724  4.034012497  4.033297194
#> wt          -2.691942008 -2.69193269 -2.691920297 -2.691905051 -2.691887157
#>                   0.0361       0.0337       0.0314      0.0293       0.0273
#> (Intercept) 21.842825221 10.577042167 10.575254770 10.57350480 10.571791799
#> am           1.158110952  1.033362600  1.033108460  1.03285897  1.032614120
#> carb         0.326654895  0.306067519  0.305904408  0.30574543  0.305590520
#> cyl          0.929746434  0.962838290  0.962567543  0.96230773  0.962058536
#> disp         0.006420906  0.007419515  0.007418238  0.00741696  0.007415684
#> drat         0.000000000  0.361531764  0.361666844  0.36179971  0.361930356
#> gear         1.224295981  1.496546018  1.496727485  1.49690572  1.497080735
#> hp          -0.052145663 -0.045380253 -0.045374161 -0.04536826 -0.045362554
#> qsec         0.000000000  0.535675912  0.535824508  0.53596774  0.536105777
#> vs           4.032603447  2.950662793  2.949455879  2.94829076  2.947166316
#> wt          -2.691866813 -3.114882750 -3.114829991 -3.11477511 -3.114718278
#>                   0.0255       0.0237       0.0221       0.0207       0.0193
#> (Intercept) 10.570115279 10.568474773 10.566869806 10.565299905 10.563764596
#> am           1.032373912  1.032138331  1.031907364  1.031680991  1.031459189
#> carb         0.305439606  0.305292621  0.305149496  0.305010162  0.304874551
#> cyl          0.961819626  0.961590688  0.961371416  0.961161507  0.960960665
#> disp         0.007414411  0.007413141  0.007411876  0.007410616  0.007409364
#> drat         0.362058768  0.362184941  0.362308873  0.362430563  0.362550011
#> gear         1.497252527  1.497421107  1.497586486  1.497748676  1.497907693
#> hp          -0.045357031 -0.045351688 -0.045346520 -0.045341524 -0.045336695
#> qsec         0.536238758  0.536366835  0.536490153  0.536608853  0.536723072
#> vs           2.946081475  2.945035166  2.944026341  2.943053970  2.942117042
#> wt          -3.114659666 -3.114599431 -3.114537726 -3.114474697 -3.114410482
#>                   0.0180       0.0168       0.0156       0.0146       0.0136
#> (Intercept) 10.562263405 10.560795857 10.559361475 10.557959785 10.556590309
#> am           1.031241933  1.031029193  1.030820938  1.030617133  1.030417739
#> carb         0.304742594  0.304614224  0.304489371  0.304367967  0.304249946
#> cyl          0.960768604  0.960585040  0.960409698  0.960242310  0.960082612
#> disp         0.007408118  0.007406881  0.007405654  0.007404435  0.007403228
#> drat         0.362667220  0.362782197  0.362894949  0.363005484  0.363113813
#> gear         1.498063554  1.498216278  1.498365886  1.498512401  1.498655848
#> hp          -0.045332029 -0.045327520 -0.045323166 -0.045318962 -0.045314903
#> qsec         0.536832947  0.536938607  0.537040181  0.537137794  0.537231568
#> vs           2.941214568  2.940345574  2.939509108  2.938704235  2.937930041
#> wt          -3.114345214 -3.114279018 -3.114212013 -3.114144314 -3.114076030
#>                   0.0127       0.0118       0.0110       0.0103       0.0096
#> (Intercept) 10.555252572 10.553946096 10.552670405 10.551425023 10.550209472
#> am           1.030222718  1.030032027  1.029845622  1.029663456  1.029485482
#> carb         0.304135240  0.304023782  0.303915505  0.303810344  0.303708233
#> cyl          0.959930348  0.959785269  0.959647130  0.959515694  0.959390728
#> disp         0.007402032  0.007400847  0.007399674  0.007398515  0.007397368
#> drat         0.363219948  0.363323903  0.363425694  0.363525336  0.363622848
#> gear         1.498796252  1.498933642  1.499068046  1.499199494  1.499328018
#> hp          -0.045310986 -0.045307208 -0.045303563 -0.045300049 -0.045296662
#> qsec         0.537321622  0.537408072  0.537491031  0.537570610  0.537646916
#> vs           2.937185627  2.936470118  2.935782653  2.935122393  2.934488513
#> wt          -3.114007262 -3.113938109 -3.113868663 -3.113799012 -3.113729239
#>                   0.0089       0.0083       0.0078       0.0073       0.0068
#> (Intercept) 10.549023278 10.547865965 10.546737058 10.545636084 10.544562572
#> am           1.029311649  1.029141908  1.028976205  1.028814488  1.028656700
#> carb         0.303609107  0.303512902  0.303419554  0.303329000  0.303241179
#> cyl          0.959272007  0.959159311  0.959052426  0.958951144  0.958855261
#> disp         0.007396235  0.007395116  0.007394012  0.007392922  0.007391847
#> drat         0.363718248  0.363811556  0.363902793  0.363991981  0.364079143
#> gear         1.499453651  1.499576425  1.499696376  1.499813538  1.499927948
#> hp          -0.045293398 -0.045290254 -0.045287226 -0.045284312 -0.045281506
#> qsec         0.537720054  0.537790126  0.537857232  0.537921468  0.537982929
#> vs           2.933880211  2.933296700  2.932737212  2.932200995  2.931687317
#> wt          -3.113659422 -3.113589636 -3.113519950 -3.113450432 -3.113381142
#>                   0.0063       0.0059       0.0055
#> (Intercept) 10.543516050 10.542496049 10.541502101
#> am           1.028502786  1.028352691  1.028206355
#> carb         0.303156027  0.303073484  0.302993490
#> cyl          0.958764579  0.958678908  0.958598061
#> disp         0.007390788  0.007389743  0.007388715
#> drat         0.364164301  0.364247481  0.364328706
#> gear         1.500039642  1.500148659  1.500255035
#> hp          -0.045278808 -0.045276212 -0.045273717
#> qsec         0.538041706  0.538097890  0.538151568
#> vs           2.931195461  2.930724728  2.930274436
#> wt          -3.113312140 -3.113243480 -3.113175215
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1  78 100  26   1   1   1   1   1   1   1   9   8   8
#>  [37]   8   8   8   8   8   8   6  86   1   1   1   1   1   1   1   1  39   1
#>  [55]   1   1   1   1   1 202   1   1   1   1   1 166   1   1   1   1   1   1
#>  [73]   1 255   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.485186320 5.115503120 4.770735331 4.449203736 4.149342296 3.869690512
#>   [7] 3.608886323 3.365659463 3.138825280 2.927278962 2.729990157 2.545997957
#>  [13] 2.374406216 2.214379184 2.065137438 1.925954087 1.796151227 1.675096646
#>  [19] 1.562200738 1.456913637 1.358722534 1.267149183 1.181747569 1.102101738
#>  [25] 1.027823769 0.958551887 0.893948698 0.833699548 0.777510989 0.725109350
#>  [31] 0.676239407 0.630663134 0.588158550 0.548518632 0.511550312 0.477073533
#>  [37] 0.444920374 0.414934230 0.386969051 0.360888632 0.336565946 0.313882528
#>  [43] 0.292727896 0.272999016 0.254599796 0.237440622 0.221437918 0.206513744
#>  [49] 0.192595409 0.179615123 0.167509665 0.156220074 0.145691364 0.135872253
#>  [55] 0.126714918 0.118174757 0.110210174 0.102782377 0.095855189 0.089394870
#>  [61] 0.083369955 0.077751099 0.072510935 0.067623941 0.063066313 0.058815855
#>  [67] 0.054851863 0.051155031 0.047707353 0.044492037 0.041493423 0.038696905
#>  [73] 0.036088863 0.033656595 0.031388253 0.029272790 0.027299902 0.025459980
#>  [79] 0.023744062 0.022143792 0.020651374 0.019259541 0.017961512 0.016750966
#>  [85] 0.015622007 0.014569136 0.013587225 0.012671492 0.011817476 0.011021017
#>  [91] 0.010278238 0.009585519 0.008939487 0.008336995 0.007775110 0.007251094
#>  [97] 0.006762394 0.006306631 0.005881585 0.005485186
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 23
#> 
#> $loss
#>   [1] 826.44667 735.85698 633.37462 547.92189 477.03257 418.57837 370.72383
#>   [8] 331.88709 300.70589 276.00805 256.78589 242.17395 231.42978 223.91720
#>  [15] 219.09179 216.48841 215.71013 215.71013 215.71013 215.71013 215.71013
#>  [22] 215.71013 215.71013 158.14176 149.47069 139.34334 139.34333 139.34333
#>  [29] 139.34333 139.34333 139.34333 139.34333 139.34333 136.88106 134.63584
#>  [36] 132.85695 131.47347 130.42289 129.65156 129.11351 128.76941 128.58564
#>  [43] 128.53341 105.52557 105.52557 105.52556 105.52556 105.52556 105.52556
#>  [50] 105.52556 105.52556 105.52556 103.35086 103.35086 103.35085 103.35085
#>  [57] 103.35085 103.35085 103.35085  99.32851  99.32850  99.32849  99.32848
#>  [64]  99.32848  99.32847  98.27880  98.27879  98.27878  98.27877  98.27876
#>  [71]  98.27875  98.27874  98.27873  97.55240  97.55237  97.55235  97.55233
#>  [78]  97.55232  97.55230  97.55228  97.55227  97.55226  97.55224  97.55223
#>  [85]  97.55222  97.55221  97.55220  97.55219  97.55218  97.55217  97.55217
#>  [92]  97.55216  97.55215  97.55215  97.55214  97.55214  97.55213  97.55213
#>  [99]  97.55213  97.55212
#> 
#> $linear.predictors
#>      5.4852   5.1155   4.7707   4.4492   4.1493   3.8697   3.6089   3.3657
#> 1  19.76667 20.27223 20.90072 21.48685 22.03347 22.54326 23.01869 23.46207
#> 2  19.76667 19.82986 19.90842 19.98169 20.05002 20.11374 20.17317 20.22859
#> 3  19.76667 19.38750 18.91613 18.47653 18.06656 17.68422 17.32765 16.99511
#> 4  19.76667 19.38750 18.91613 18.47653 18.06656 17.68422 17.32765 16.99511
#> 5  19.76667 20.27223 20.90072 21.48685 22.03347 22.54326 23.01869 23.46207
#> 6  19.76667 19.38750 18.91613 18.47653 18.06656 17.68422 17.32765 16.99511
#> 7  19.76667 19.38750 18.91613 18.47653 18.06656 17.68422 17.32765 16.99511
#> 8  19.76667 19.38750 18.91613 18.47653 18.06656 17.68422 17.32765 16.99511
#> 9  19.76667 19.38750 18.91613 18.47653 18.06656 17.68422 17.32765 16.99511
#> 10 19.76667 20.27223 20.90072 21.48685 22.03347 22.54326 23.01869 23.46207
#> 11 19.76667 20.27223 20.90072 21.48685 22.03347 22.54326 23.01869 23.46207
#> 12 19.76667 20.27223 20.90072 21.48685 22.03347 22.54326 23.01869 23.46207
#> 13 19.76667 19.38750 18.91613 18.47653 18.06656 17.68422 17.32765 16.99511
#> 14 19.76667 19.38750 18.91613 18.47653 18.06656 17.68422 17.32765 16.99511
#> 15 19.76667 19.38750 18.91613 18.47653 18.06656 17.68422 17.32765 16.99511
#> 16 19.76667 20.27223 20.90072 21.48685 22.03347 22.54326 23.01869 23.46207
#> 17 19.76667 20.27223 20.90072 21.48685 22.03347 22.54326 23.01869 23.46207
#> 18 19.76667 19.38750 18.91613 18.47653 18.06656 17.68422 17.32765 16.99511
#> 19 19.76667 19.82986 19.90842 19.98169 20.05002 20.11374 20.17317 20.22859
#> 20 19.76667 19.38750 18.91613 18.47653 18.06656 17.68422 17.32765 16.99511
#> 21 19.76667 20.27223 20.90072 21.48685 22.03347 22.54326 23.01869 23.46207
#>      3.1388   2.9273   2.7300   2.5460   2.3744   2.2144   2.0651   1.9260
#> 1  23.87557 24.26121 24.62085 24.95626 25.26906 25.56078 25.83283 26.08655
#> 2  20.28028 20.32848 20.37344 20.41537 20.45447 20.49093 20.52494 20.55665
#> 3  16.68499 16.39576 16.12603 15.87447 15.63987 15.42108 15.21704 15.02675
#> 4  16.68499 16.39576 16.12603 15.87447 15.63987 15.42108 15.21704 15.02675
#> 5  23.87557 24.26121 24.62085 24.95626 25.26906 25.56078 25.83283 26.08655
#> 6  16.68499 16.39576 16.12603 15.87447 15.63987 15.42108 15.21704 15.02675
#> 7  16.68499 16.39576 16.12603 15.87447 15.63987 15.42108 15.21704 15.02675
#> 8  16.68499 16.39576 16.12603 15.87447 15.63987 15.42108 15.21704 15.02675
#> 9  16.68499 16.39576 16.12603 15.87447 15.63987 15.42108 15.21704 15.02675
#> 10 23.87557 24.26121 24.62085 24.95626 25.26906 25.56078 25.83283 26.08655
#> 11 23.87557 24.26121 24.62085 24.95626 25.26906 25.56078 25.83283 26.08655
#> 12 23.87557 24.26121 24.62085 24.95626 25.26906 25.56078 25.83283 26.08655
#> 13 16.68499 16.39576 16.12603 15.87447 15.63987 15.42108 15.21704 15.02675
#> 14 16.68499 16.39576 16.12603 15.87447 15.63987 15.42108 15.21704 15.02675
#> 15 16.68499 16.39576 16.12603 15.87447 15.63987 15.42108 15.21704 15.02675
#> 16 23.87557 24.26121 24.62085 24.95626 25.26906 25.56078 25.83283 26.08655
#> 17 23.87557 24.26121 24.62085 24.95626 25.26906 25.56078 25.83283 26.08655
#> 18 16.68499 16.39576 16.12603 15.87447 15.63987 15.42108 15.21704 15.02675
#> 19 20.28028 20.32848 20.37344 20.41537 20.45447 20.49093 20.52494 20.55665
#> 20 16.68499 16.39576 16.12603 15.87447 15.63987 15.42108 15.21704 15.02675
#> 21 23.87557 24.26121 24.62085 24.95626 25.26906 25.56078 25.83283 26.08655
#>      1.7962   1.6751   1.5622   1.4569   1.3587   1.2671   1.1817   1.1021
#> 1  26.32051 26.32051 26.32051 26.32051 26.32051 26.32051 26.32051 26.19563
#> 2  20.58590 20.58590 20.58590 20.58590 20.58590 20.58590 20.58590 20.49652
#> 3  14.85128 14.85128 14.85128 14.85128 14.85128 14.85128 14.85128 15.75550
#> 4  14.85128 14.85128 14.85128 14.85128 14.85128 14.85128 14.85128 15.56960
#> 5  26.32051 26.32051 26.32051 26.32051 26.32051 26.32051 26.32051 24.95155
#> 6  14.85128 14.85128 14.85128 14.85128 14.85128 14.85128 14.85128 14.85462
#> 7  14.85128 14.85128 14.85128 14.85128 14.85128 14.85128 14.85128 13.16725
#> 8  14.85128 14.85128 14.85128 14.85128 14.85128 14.85128 14.85128 12.91844
#> 9  14.85128 14.85128 14.85128 14.85128 14.85128 14.85128 14.85128 13.03140
#> 10 26.32051 26.32051 26.32051 26.32051 26.32051 26.32051 26.32051 26.36722
#> 11 26.32051 26.32051 26.32051 26.32051 26.32051 26.32051 26.32051 27.20376
#> 12 26.32051 26.32051 26.32051 26.32051 26.32051 26.32051 26.32051 25.98828
#> 13 14.85128 14.85128 14.85128 14.85128 14.85128 14.85128 14.85128 15.76265
#> 14 14.85128 14.85128 14.85128 14.85128 14.85128 14.85128 14.85128 15.18351
#> 15 14.85128 14.85128 14.85128 14.85128 14.85128 14.85128 14.85128 15.17636
#> 16 26.32051 26.32051 26.32051 26.32051 26.32051 26.32051 26.32051 26.74617
#> 17 26.32051 26.32051 26.32051 26.32051 26.32051 26.32051 26.32051 27.34962
#> 18 14.85128 14.85128 14.85128 14.85128 14.85128 14.85128 14.85128 16.14159
#> 19 20.58590 20.58590 20.58590 20.58590 20.58590 20.58590 20.58590 21.13286
#> 20 14.85128 14.85128 14.85128 14.85128 14.85128 14.85128 14.85128 15.56960
#> 21 26.32051 26.32051 26.32051 26.32051 26.32051 26.32051 26.32051 25.53784
#>      1.0278   0.9586   0.8939   0.8337   0.7775   0.7251   0.6762   0.6307
#> 1  25.48227 26.07112 26.07126 26.07135 26.07140 26.07144 26.07146 26.07147
#> 2  20.29261 20.40830 20.40833 20.40834 20.40836 20.40836 20.40837 20.40837
#> 3  17.55139 16.64732 16.64711 16.64698 16.64689 16.64684 16.64681 16.64679
#> 4  17.07632 16.27831 16.27812 16.27800 16.27793 16.27788 16.27786 16.27784
#> 5  22.30295 23.60156 23.60187 23.60206 23.60218 23.60225 23.60230 23.60233
#> 6  15.24913 14.85902 14.85893 14.85887 14.85884 14.85881 14.85880 14.85879
#> 7  10.93696 11.50951 11.50964 11.50973 11.50978 11.50981 11.50983 11.50984
#> 8  10.30109 11.01559 11.01576 11.01587 11.01593 11.01597 11.01600 11.01602
#> 9  10.58979 11.23984 11.23999 11.24009 11.24015 11.24019 11.24021 11.24023
#> 10 25.92079 26.41175 26.41187 26.41194 26.41198 26.41201 26.41203 26.41204
#> 11 28.05861 28.07232 28.07232 28.07232 28.07232 28.07232 28.07232 28.07232
#> 12 24.95238 25.65953 25.65970 25.65980 25.65986 25.65991 25.65993 25.65995
#> 13 17.56966 16.66152 16.66130 16.66117 16.66108 16.66103 16.66100 16.66098
#> 14 16.08964 15.51189 15.51176 15.51167 15.51162 15.51159 15.51157 15.51155
#> 15 16.07137 15.49770 15.49757 15.49748 15.49743 15.49740 15.49738 15.49736
#> 16 26.88921 27.16397 27.16404 27.16408 27.16410 27.16412 27.16413 27.16413
#> 17 28.43136 28.36185 28.36183 28.36182 28.36182 28.36181 28.36181 28.36181
#> 18 18.53808 17.41374 17.41347 17.41331 17.41320 17.41314 17.41310 17.41307
#> 19 21.91881 21.67147 21.67141 21.67137 21.67135 21.67133 21.67132 21.67132
#> 20 17.07632 16.27831 16.27812 16.27800 16.27793 16.27788 16.27786 16.27784
#> 21 23.80125 24.76538 24.76561 24.76575 24.76584 24.76589 24.76593 24.76595
#>      0.5882   0.5485   0.5116   0.4771   0.4449   0.4149   0.3870   0.3609
#> 1  26.07148 26.10742 26.14512 26.18029 26.21310 26.24369 26.27222 26.29883
#> 2  20.40837 20.51033 20.61747 20.71746 20.81072 20.89769 20.97880 21.05444
#> 3  16.64678 16.75582 16.87053 16.97761 17.07747 17.17061 17.25747 17.33847
#> 4  16.27783 16.26659 16.25489 16.24398 16.23380 16.22432 16.21547 16.20721
#> 5  23.60235 23.57683 23.54985 23.52465 23.50115 23.47923 23.45878 23.43972
#> 6  14.85879 14.90677 14.95724 15.00435 15.04829 15.08927 15.12749 15.16313
#> 7  11.50985 11.49587 11.48109 11.46730 11.45443 11.44243 11.43124 11.42081
#> 8  11.01603 11.00175 10.98667 10.97257 10.95943 10.94717 10.93574 10.92507
#> 9  11.24023 11.22609 11.21115 11.19719 11.18417 11.17203 11.16071 11.15015
#> 10 26.41205 26.44819 26.48610 26.52148 26.55448 26.58525 26.61395 26.64072
#> 11 28.07232 28.04939 28.02529 28.00279 27.98181 27.96224 27.94399 27.92697
#> 12 25.65996 25.69566 25.73309 25.76802 25.80059 25.83097 25.85930 25.88572
#> 13 16.66097 16.77002 16.88474 16.99183 17.09170 17.18484 17.27171 17.35272
#> 14 15.51154 15.49987 15.48767 15.47630 15.46569 15.45580 15.44657 15.43797
#> 15 15.49735 15.60573 15.71970 15.82609 15.92531 16.01783 16.10413 16.18460
#> 16 27.16414 27.20071 27.23911 27.27495 27.30837 27.33954 27.36861 27.39572
#> 17 28.36181 28.33904 28.31512 28.29280 28.27198 28.25257 28.23446 28.21758
#> 18 17.41306 17.40248 17.39151 17.38129 17.37175 17.36286 17.35457 17.34683
#> 19 21.67132 21.47385 21.26636 21.07270 20.89209 20.72365 20.56656 20.42006
#> 20 16.27783 16.02647 15.76241 15.51597 15.28613 15.07178 14.87188 14.68545
#> 21 24.76596 24.74112 24.71489 24.69039 24.66754 24.64623 24.62636 24.60783
#>      0.3366   0.3139   0.2927   0.2730   0.2546   0.2374   0.2214   0.2065
#> 1  26.32365 26.34679 26.36729 25.53344 25.53334 25.53327 25.53322 25.53318
#> 2  21.12499 21.19078 21.24896 21.12370 21.12375 21.12379 21.12381 21.12383
#> 3  17.41401 17.48447 17.54671 17.72979 17.73007 17.73027 17.73041 17.73051
#> 4  16.19951 16.19234 16.18594 14.34494 14.34551 14.34592 14.34621 14.34641
#> 5  23.42194 23.40536 23.39077 24.80139 24.80081 24.80040 24.80012 24.79991
#> 6  15.19637 15.22737 15.25475 15.99503 15.99510 15.99515 15.99519 15.99521
#> 7  11.41108 11.40200 11.39401 12.05962 12.05937 12.05919 12.05907 12.05898
#> 8  10.91513 10.90586 10.89770 11.20222 11.20196 11.20177 11.20163 11.20154
#> 9  11.14030 11.13112 11.12304 10.73392 10.73375 10.73364 10.73356 10.73350
#> 10 26.66568 26.68896 26.70957 27.00705 27.00686 27.00673 27.00664 27.00657
#> 11 27.91110 27.89630 27.88321 29.02856 29.02852 29.02850 29.02848 29.02847
#> 12 25.91036 25.93334 25.95370 23.49819 23.49846 23.49865 23.49879 23.49888
#> 13 17.42827 17.49872 17.56097 18.83902 18.83918 18.83930 18.83938 18.83944
#> 14 15.42994 15.42246 15.41581 13.69552 13.69599 13.69633 13.69656 13.69673
#> 15 16.25966 16.32965 16.39151 16.75566 16.75579 16.75588 16.75595 16.75599
#> 16 27.42100 27.44458 27.46544 27.64444 27.64435 27.64429 27.64425 27.64422
#> 17 28.20183 28.18714 28.17415 28.10765 28.10755 28.10747 28.10742 28.10739
#> 18 17.33962 17.33289 17.32688 17.49503 17.49504 17.49505 17.49505 17.49505
#> 19 20.28344 20.15602 20.04334 22.36318 22.36290 22.36271 22.36257 22.36247
#> 20 14.51158 14.34944 14.20601 13.41688 13.41708 13.41722 13.41732 13.41739
#> 21 24.59055 24.57443 24.56023 23.72481 23.72461 23.72447 23.72438 23.72431
#>      0.1926   0.1796   0.1675   0.1562   0.1457   0.1359   0.1267   0.1182
#> 1  25.53316 25.53314 25.53313 25.53312 25.73722 25.73713 25.73705 25.73699
#> 2  21.12384 21.12385 21.12386 21.12386 21.66900 21.66901 21.66900 21.66899
#> 3  17.73058 17.73063 17.73067 17.73069 17.31779 17.31805 17.31826 17.31842
#> 4  14.34655 14.34665 14.34672 14.34677 14.41556 14.41606 14.41642 14.41668
#> 5  24.79977 24.79967 24.79960 24.79955 24.90829 24.90777 24.90740 24.90712
#> 6  15.99523 15.99525 15.99525 15.99526 15.69570 15.69577 15.69583 15.69588
#> 7  12.05892 12.05887 12.05884 12.05882 12.07544 12.07522 12.07506 12.07494
#> 8  11.20147 11.20143 11.20140 11.20137 11.30657 11.30633 11.30615 11.30602
#> 9  10.73346 10.73343 10.73341 10.73339 10.92822 10.92807 10.92795 10.92786
#> 10 27.00653 27.00649 27.00647 27.00645 27.01588 27.01571 27.01559 27.01551
#> 11 29.02846 29.02846 29.02845 29.02845 28.87045 28.87043 28.87041 28.87041
#> 12 23.49895 23.49900 23.49903 23.49905 23.86243 23.86264 23.86278 23.86288
#> 13 18.83948 18.83951 18.83953 18.83954 18.25988 18.26004 18.26017 18.26029
#> 14 13.69685 13.69693 13.69699 13.69703 13.80015 13.80056 13.80086 13.80107
#> 15 16.75603 16.75605 16.75607 16.75608 16.39468 16.39480 16.39491 16.39499
#> 16 27.64420 27.64418 27.64417 27.64417 27.61989 27.61982 27.61976 27.61972
#> 17 28.10736 28.10734 28.10733 28.10732 28.22745 28.22736 28.22730 28.22724
#> 18 17.49505 17.49505 17.49505 17.49506 17.41073 17.41076 17.41078 17.41080
#> 19 22.36241 22.36236 22.36232 22.36230 21.63570 21.63550 21.63538 21.63531
#> 20 13.41744 13.41748 13.41750 13.41752 13.85586 13.85605 13.85616 13.85623
#> 21 23.72426 23.72423 23.72420 23.72419 24.09311 24.09292 24.09277 24.09266
#>      0.1102   0.1028   0.0959   0.0894   0.0834   0.0778   0.0725   0.0676
#> 1  25.73694 25.73691 25.73688 25.45694 25.45688 25.45683 25.45678 25.45674
#> 2  21.66897 21.66894 21.66892 22.45095 22.45130 22.45164 22.45196 22.45227
#> 3  17.31854 17.31863 17.31870 17.52079 17.52083 17.52087 17.52090 17.52093
#> 4  14.41686 14.41699 14.41707 14.05269 14.05266 14.05264 14.05263 14.05261
#> 5  24.90693 24.90679 24.90670 24.86275 24.86270 24.86264 24.86259 24.86254
#> 6  15.69593 15.69597 15.69600 15.57313 15.57320 15.57326 15.57331 15.57336
#> 7  12.07486 12.07480 12.07476 12.10751 12.10745 12.10739 12.10734 12.10729
#> 8  11.30593 11.30586 11.30581 11.25635 11.25629 11.25623 11.25617 11.25612
#> 9  10.92780 10.92775 10.92771 10.83448 10.83441 10.83435 10.83429 10.83423
#> 10 27.01544 27.01540 27.01538 27.02721 27.02719 27.02717 27.02716 27.02714
#> 11 28.87041 28.87041 28.87041 29.50145 29.50145 29.50144 29.50144 29.50143
#> 12 23.86294 23.86298 23.86300 23.01349 23.01333 23.01317 23.01303 23.01289
#> 13 18.26038 18.26046 18.26052 18.53444 18.53452 18.53460 18.53467 18.53474
#> 14 13.80121 13.80131 13.80138 13.69779 13.69777 13.69774 13.69772 13.69770
#> 15 16.39506 16.39511 16.39516 16.64227 16.64228 16.64230 16.64231 16.64233
#> 16 27.61970 27.61968 27.61967 27.64263 27.64261 27.64259 27.64258 27.64257
#> 17 28.22720 28.22717 28.22714 28.20527 28.20532 28.20536 28.20540 28.20544
#> 18 17.41081 17.41081 17.41081 18.40334 18.40350 18.40366 18.40380 18.40394
#> 19 21.63527 21.63526 21.63526 20.76187 20.76161 20.76136 20.76113 20.76090
#> 20 13.85627 13.85628 13.85628 13.71525 13.71538 13.71550 13.71561 13.71572
#> 21 24.09257 24.09250 24.09245 23.83940 23.83932 23.83925 23.83917 23.83910
#>      0.0631   0.0588   0.0549   0.0512   0.0477   0.0445   0.0415   0.0387
#> 1  25.45669 25.57463 25.57471 25.57479 25.57486 25.57493 25.57501 25.57508
#> 2  22.45256 22.72179 22.72146 22.72114 22.72083 22.72053 22.72023 22.71995
#> 3  17.52097 17.53297 17.53298 17.53299 17.53300 17.53300 17.53300 17.53301
#> 4  14.05259 14.18589 14.18594 14.18599 14.18603 14.18607 14.18611 14.18615
#> 5  24.86249 24.26647 24.26651 24.26655 24.26659 24.26663 24.26668 24.26672
#> 6  15.57341 15.36291 15.36282 15.36274 15.36266 15.36259 15.36253 15.36247
#> 7  12.10725 12.46872 12.46874 12.46875 12.46877 12.46878 12.46879 12.46880
#> 8  11.25606 11.40168 11.40172 11.40176 11.40179 11.40183 11.40186 11.40189
#> 9  10.83418 10.70356 10.70362 10.70368 10.70374 10.70380 10.70386 10.70392
#> 10 27.02712 27.11797 27.11797 27.11797 27.11796 27.11796 27.11797 27.11797
#> 11 29.50143 29.73085 29.73075 29.73065 29.73055 29.73044 29.73034 29.73024
#> 12 23.01276 22.67023 22.67043 22.67063 22.67082 22.67100 22.67118 22.67136
#> 13 18.53481 18.49105 18.49097 18.49089 18.49081 18.49074 18.49067 18.49060
#> 14 13.69768 13.39488 13.39492 13.39497 13.39501 13.39505 13.39509 13.39513
#> 15 16.64234 16.69947 16.69950 16.69952 16.69955 16.69958 16.69960 16.69962
#> 16 27.64255 27.83326 27.83326 27.83326 27.83325 27.83325 27.83325 27.83324
#> 17 28.20548 28.17253 28.17256 28.17258 28.17260 28.17263 28.17265 28.17267
#> 18 18.40408 17.82135 17.82131 17.82127 17.82123 17.82119 17.82115 17.82111
#> 19 20.76069 21.00898 21.00912 21.00925 21.00939 21.00952 21.00965 21.00977
#> 20 13.71582 14.02854 14.02839 14.02824 14.02810 14.02796 14.02782 14.02769
#> 21 23.83904 23.91227 23.91234 23.91240 23.91246 23.91251 23.91257 23.91263
#>      0.0361   0.0337   0.0314   0.0293   0.0273   0.0255   0.0237   0.0221
#> 1  25.57514 25.41991 25.41998 25.42005 25.42011 25.42017 25.42023 25.42028
#> 2  22.71967 22.53556 22.53513 22.53473 22.53433 22.53395 22.53358 22.53322
#> 3  17.53301 17.65184 17.65184 17.65185 17.65185 17.65186 17.65186 17.65186
#> 4  14.18618 14.07201 14.07196 14.07191 14.07185 14.07180 14.07175 14.07169
#> 5  24.26677 24.36334 24.36349 24.36364 24.36379 24.36393 24.36407 24.36421
#> 6  15.36242 15.31854 15.31860 15.31866 15.31872 15.31879 15.31885 15.31892
#> 7  12.46881 12.53032 12.53025 12.53018 12.53011 12.53004 12.52997 12.52990
#> 8  11.40192 11.38510 11.38510 11.38510 11.38509 11.38509 11.38509 11.38509
#> 9  10.70397 10.67096 10.67104 10.67113 10.67121 10.67129 10.67136 10.67144
#> 10 27.11797 27.34541 27.34550 27.34559 27.34568 27.34577 27.34585 27.34593
#> 11 29.73014 29.88516 29.88495 29.88475 29.88455 29.88436 29.88418 29.88400
#> 12 22.67153 23.04232 23.04266 23.04300 23.04332 23.04363 23.04393 23.04422
#> 13 18.49054 18.53641 18.53638 18.53635 18.53632 18.53630 18.53627 18.53625
#> 14 13.39517 13.11445 13.11443 13.11442 13.11440 13.11438 13.11437 13.11436
#> 15 16.69964 16.67785 16.67782 16.67780 16.67777 16.67774 16.67771 16.67768
#> 16 27.83324 27.86774 27.86774 27.86773 27.86773 27.86772 27.86771 27.86771
#> 17 28.17269 27.78799 27.78791 27.78783 27.78775 27.78767 27.78760 27.78753
#> 18 17.82107 18.06276 18.06286 18.06296 18.06305 18.06315 18.06324 18.06333
#> 19 21.00989 20.82435 20.82443 20.82451 20.82459 20.82466 20.82473 20.82480
#> 20 14.02756 14.25583 14.25572 14.25561 14.25550 14.25540 14.25530 14.25520
#> 21 23.91268 23.75215 23.75219 23.75223 23.75227 23.75231 23.75235 23.75239
#>      0.0207   0.0193   0.0180   0.0168   0.0156   0.0146   0.0136   0.0127
#> 1  25.42034 25.42039 25.42044 25.42049 25.42053 25.42057 25.42062 25.42066
#> 2  22.53288 22.53254 22.53222 22.53191 22.53161 22.53132 22.53104 22.53078
#> 3  17.65186 17.65186 17.65186 17.65185 17.65185 17.65185 17.65184 17.65184
#> 4  14.07164 14.07159 14.07153 14.07148 14.07143 14.07137 14.07132 14.07127
#> 5  24.36435 24.36448 24.36462 24.36475 24.36487 24.36500 24.36512 24.36524
#> 6  15.31899 15.31905 15.31912 15.31919 15.31926 15.31933 15.31940 15.31946
#> 7  12.52983 12.52977 12.52970 12.52963 12.52957 12.52950 12.52943 12.52937
#> 8  11.38509 11.38508 11.38508 11.38507 11.38507 11.38506 11.38506 11.38505
#> 9  10.67151 10.67158 10.67165 10.67172 10.67178 10.67184 10.67191 10.67197
#> 10 27.34601 27.34608 27.34616 27.34623 27.34630 27.34636 27.34643 27.34649
#> 11 29.88382 29.88365 29.88349 29.88333 29.88318 29.88303 29.88289 29.88275
#> 12 23.04450 23.04477 23.04503 23.04528 23.04552 23.04575 23.04598 23.04619
#> 13 18.53623 18.53621 18.53619 18.53617 18.53616 18.53614 18.53613 18.53612
#> 14 13.11435 13.11434 13.11433 13.11432 13.11431 13.11430 13.11430 13.11429
#> 15 16.67765 16.67762 16.67759 16.67756 16.67753 16.67750 16.67747 16.67744
#> 16 27.86770 27.86769 27.86769 27.86768 27.86767 27.86766 27.86766 27.86765
#> 17 27.78745 27.78739 27.78732 27.78725 27.78719 27.78713 27.78707 27.78701
#> 18 18.06342 18.06351 18.06360 18.06368 18.06376 18.06384 18.06392 18.06400
#> 19 20.82487 20.82493 20.82499 20.82505 20.82511 20.82516 20.82521 20.82526
#> 20 14.25510 14.25501 14.25492 14.25483 14.25474 14.25466 14.25458 14.25450
#> 21 23.75243 23.75247 23.75250 23.75253 23.75257 23.75260 23.75263 23.75266
#>      0.0118   0.0110   0.0103   0.0096   0.0089   0.0083   0.0078   0.0073
#> 1  25.42069 25.42073 25.42076 25.42080 25.42083 25.42086 25.42089 25.42092
#> 2  22.53052 22.53027 22.53003 22.52980 22.52958 22.52936 22.52916 22.52896
#> 3  17.65183 17.65182 17.65181 17.65181 17.65180 17.65179 17.65178 17.65177
#> 4  14.07122 14.07117 14.07112 14.07107 14.07102 14.07097 14.07092 14.07087
#> 5  24.36536 24.36548 24.36559 24.36570 24.36581 24.36592 24.36602 24.36612
#> 6  15.31953 15.31960 15.31967 15.31974 15.31981 15.31988 15.31994 15.32001
#> 7  12.52931 12.52924 12.52918 12.52912 12.52906 12.52900 12.52894 12.52889
#> 8  11.38505 11.38504 11.38504 11.38503 11.38502 11.38501 11.38501 11.38500
#> 9  10.67202 10.67208 10.67214 10.67219 10.67224 10.67229 10.67234 10.67239
#> 10 27.34655 27.34661 27.34667 27.34672 27.34678 27.34683 27.34688 27.34693
#> 11 29.88262 29.88249 29.88236 29.88224 29.88212 29.88201 29.88190 29.88179
#> 12 23.04640 23.04660 23.04679 23.04698 23.04716 23.04733 23.04749 23.04765
#> 13 18.53611 18.53610 18.53609 18.53608 18.53607 18.53607 18.53606 18.53606
#> 14 13.11429 13.11429 13.11428 13.11428 13.11428 13.11428 13.11428 13.11428
#> 15 16.67741 16.67738 16.67735 16.67732 16.67729 16.67726 16.67723 16.67720
#> 16 27.86764 27.86764 27.86763 27.86762 27.86761 27.86761 27.86760 27.86759
#> 17 27.78695 27.78689 27.78684 27.78679 27.78674 27.78669 27.78664 27.78659
#> 18 18.06408 18.06415 18.06423 18.06430 18.06437 18.06444 18.06450 18.06457
#> 19 20.82531 20.82536 20.82540 20.82544 20.82548 20.82552 20.82556 20.82559
#> 20 14.25442 14.25435 14.25428 14.25421 14.25414 14.25408 14.25401 14.25395
#> 21 23.75269 23.75272 23.75275 23.75278 23.75280 23.75283 23.75285 23.75288
#>      0.0068   0.0063   0.0059   0.0055
#> 1  25.42094 25.42097 25.42099 25.42101
#> 2  22.52877 22.52859 22.52841 22.52824
#> 3  17.65176 17.65175 17.65174 17.65173
#> 4  14.07082 14.07078 14.07073 14.07068
#> 5  24.36622 24.36632 24.36641 24.36651
#> 6  15.32008 15.32014 15.32021 15.32027
#> 7  12.52883 12.52877 12.52872 12.52867
#> 8  11.38499 11.38499 11.38498 11.38497
#> 9  10.67244 10.67248 10.67252 10.67257
#> 10 27.34697 27.34702 27.34706 27.34710
#> 11 29.88169 29.88159 29.88150 29.88140
#> 12 23.04780 23.04794 23.04808 23.04821
#> 13 18.53606 18.53605 18.53605 18.53605
#> 14 13.11428 13.11428 13.11428 13.11428
#> 15 16.67717 16.67714 16.67711 16.67708
#> 16 27.86758 27.86758 27.86757 27.86756
#> 17 27.78655 27.78650 27.78646 27.78642
#> 18 18.06463 18.06470 18.06476 18.06482
#> 19 20.82563 20.82566 20.82569 20.82572
#> 20 14.25389 14.25383 14.25378 14.25373
#> 21 23.75290 23.75292 23.75294 23.75296
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 22.8 21.4 18.7 14.3 24.4 16.4 10.4 10.4 14.7 32.4 30.4 21.5 15.2 13.3 19.2
#> [16] 27.3 30.4 15.8 19.7 15.0 21.4
#> 
#> $X
#>               am       carb        cyl        disp        drat       gear
#>  [1,]  1.1547005 -1.0430642 -1.2152872 -1.05870950  0.53253493  0.4291975
#>  [2,] -0.8660254 -1.0430642 -0.1519109  0.04080417 -0.99696593 -0.8583951
#>  [3,] -0.8660254 -0.4814142  0.9114654  0.78847347 -0.85792040 -0.8583951
#>  [4,] -0.8660254  0.6418856  0.9114654  0.78847347 -0.73873851 -0.8583951
#>  [5,] -0.8660254 -0.4814142 -1.2152872 -0.77503497  0.21471657  0.4291975
#>  [6,] -0.8660254  0.0802357  0.9114654  0.17127980 -1.01682958 -0.8583951
#>  [7,] -0.8660254  0.6418856  0.9114654  1.60944368 -1.29492064 -0.8583951
#>  [8,] -0.8660254  0.6418856  0.9114654  1.52148258 -1.15587511 -0.8583951
#>  [9,] -0.8660254  0.6418856  0.9114654  1.37488076 -0.69901122 -0.8583951
#> [10,]  1.1547005 -1.0430642 -1.2152872 -1.27348117  0.98939883  0.4291975
#> [11,]  1.1547005 -0.4814142 -1.2152872 -1.29547144  2.67780887  0.4291975
#> [12,] -0.8660254 -1.0430642 -1.2152872 -0.97001539  0.23458022 -0.8583951
#> [13,] -0.8660254 -0.4814142  0.9114654  0.37798837 -0.85792040 -0.8583951
#> [14,] -0.8660254  0.6418856  0.9114654  0.71517256  0.29417116 -0.8583951
#> [15,] -0.8660254 -0.4814142  0.9114654  1.08167712 -0.99696593 -0.8583951
#> [16,]  1.1547005 -1.0430642 -1.2152872 -1.27128214  0.98939883  0.4291975
#> [17,]  1.1547005 -0.4814142 -1.2152872 -1.15326767  0.37362575  1.7167902
#> [18,]  1.1547005  0.6418856  0.9114654  0.72250265  1.26748989  1.7167902
#> [19,]  1.1547005  1.7651855 -0.1519109 -0.78749612  0.07567104  1.7167902
#> [20,]  1.1547005  2.8884854  0.9114654  0.35599809 -0.08323814  1.7167902
#> [21,]  1.1547005 -0.4814142 -1.2152872 -0.96341831  1.04898977  0.4291975
#>               hp        qsec         vs          wt
#>  [1,] -0.8812793  0.66070131  1.1547005 -0.87980578
#>  [2,] -0.6579885  1.17207004  1.1547005 -0.06289294
#>  [3,]  0.1957704 -0.31890867 -0.8660254  0.14247621
#>  [4,]  1.1152030 -1.04591482 -0.8660254  0.26113394
#>  [5,] -1.2884566  1.51708991  1.1547005 -0.08571174
#>  [6,]  0.2614441 -0.08478805 -0.8660254  0.71750982
#>  [7,]  0.5898129  0.27255396 -0.8660254  1.79455692
#>  [8,]  0.7211604  0.17397685 -0.8660254  1.95337573
#>  [9,]  0.9181817 -0.07246591 -0.8660254  1.88126834
#> [10,] -1.2359176  1.19055325  1.1547005 -0.98933600
#> [11,] -1.4198042  0.60525169  1.1547005 -1.52329578
#> [12,] -0.8287403  1.52325098  1.1547005 -0.74745678
#> [13,] -0.1325984 -0.14639874 -0.8660254  0.13791245
#> [14,]  1.1152030 -1.31084079 -0.8660254  0.50757692
#> [15,]  0.1957704 -0.30042546 -0.8660254  0.51214067
#> [16,] -1.2359176  0.83937231  1.1547005 -1.23121522
#> [17,] -0.6185843 -0.39284150  1.1547005 -1.61639647
#> [18,]  1.3647633 -1.87149807 -0.8660254 -0.10396677
#> [19,]  0.1957704 -1.25539116 -0.8660254 -0.46906748
#> [20,]  2.2973307 -1.80988738 -0.8660254  0.26113394
#> [21,] -0.6711233  0.65454024  1.1547005 -0.45993997
#> attr(,"center")
#>  [1]   0.4285714   2.8571429   6.2857143 252.4333333   3.5819048   3.6666667
#>  [7] 160.0952381  17.5376190   0.4285714   3.2839048
#> attr(,"scale")
#>  [1]   0.4948717   1.7804685   1.8808017 136.4239517   0.5034322   0.7766432
#>  [7]  76.1339076   1.6230949   0.4948717   1.0955881
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 33
#> 
#> $lambda.min
#> [1] 0.5881585
#> 
#> $null.dev
#> [1] 39.3546
#> 
#> $Bias
#> [1] 5.514251
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 4.928535