Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 28.014076 26.158031 23.948643 22.203161 20.046661 18.187014 16.696349
#>   [8] 15.518063 14.603499 13.910878 13.404362 13.053248 12.831267 12.715978
#>  [15] 12.688243 12.638632 12.393880 12.168855 11.974645 11.807128 11.751776
#>  [22] 11.576179 11.250927 10.611950 10.133854  9.630688  8.866217  8.223733
#>  [29]  7.720203  7.561583  7.525742  7.530574  7.461895  7.436900  7.265845
#>  [36]  7.096482  6.994057  7.161636  6.802379  6.435911  6.596886  6.744760
#>  [43]  6.524020  6.849096  6.981919  6.500624  6.561138  8.383526  8.211593
#>  [50]  8.652106  8.722439  8.748078  8.920294  9.090380  9.094071  8.983202
#>  [57]  8.983475  8.955792  9.054066  8.671428  8.882489  9.231261  9.231119
#>  [64]  9.101369  9.101556  8.962205  9.054163  9.068363  8.620813  8.620762
#>  [71]  8.620722  8.697818  8.596247  8.576384  9.547522  9.629024  9.629920
#>  [78]  9.630755  9.631531  9.688494  9.689279  9.672681  9.595667  9.595883
#>  [85]  9.591050  9.591307  9.591547  9.591770  9.591909  9.506535  9.506676
#>  [92]  9.506807  9.506928  9.507039  9.507141  9.507267  9.697123  9.697287
#>  [99]  9.690907  9.692004
#> 
#> $cvse
#>   [1] 8.626010 8.406148 8.083186 7.851846 7.091566 6.334713 5.671033 5.093302
#>   [9] 4.597027 4.179875 3.840906 3.579543 3.394324 3.281724 3.235514 3.203424
#>  [17] 3.123418 3.023518 2.936883 2.870556 2.871041 2.865285 2.834424 2.782391
#>  [25] 2.750396 2.612599 2.421860 2.310115 2.239185 2.214841 2.189923 2.161680
#>  [33] 2.146911 2.152866 2.129009 2.098869 2.083076 2.074275 1.817028 1.684976
#>  [41] 1.741236 1.809925 1.812744 1.860132 1.845299 1.655545 1.640239 2.036189
#>  [49] 2.034015 2.036560 2.059037 2.054380 2.142635 2.143292 2.143317 2.154079
#>  [57] 2.154170 2.156149 2.144528 2.120445 2.132756 2.100895 2.100848 1.947131
#>  [65] 1.947134 1.919676 1.939489 1.937368 1.904552 1.904556 1.904565 1.893787
#>  [73] 1.865360 1.860628 1.919645 1.920250 1.920415 1.920571 1.920719 1.915600
#>  [81] 1.915729 1.911865 1.909603 1.909682 1.910196 1.910276 1.910352 1.910423
#>  [89] 1.910482 1.901795 1.901848 1.901898 1.901944 1.901986 1.902025 1.902087
#>  [97] 1.970567 1.970750 1.971263 1.971321
#> 
#> $fold
#>  [1]  5  2  9  9  4  3  1  6  7  8  7  4  6 10  1  1  5  8  2 10  3
#> 
#> $lambda
#>   [1] 4.524980958 4.220012386 3.935597675 3.670351564 3.422982153 3.192284612
#>   [7] 2.977135313 2.776486357 2.589360469 2.414846240 2.252093686 2.100310109
#>  [13] 1.958756237 1.826742622 1.703626283 1.588807574 1.481727261 1.381863803
#>  [19] 1.288730807 1.201874664 1.120872334 1.045329290 0.974877595 0.909174108
#>  [25] 0.847898816 0.790753273 0.737459149 0.687756871 0.641404360 0.598175852
#>  [31] 0.557860802 0.520262851 0.485198876 0.452498096 0.422001239 0.393559768
#>  [37] 0.367035156 0.342298215 0.319228461 0.297713531 0.277648636 0.258936047
#>  [43] 0.241484624 0.225209369 0.210031011 0.195875624 0.182674262 0.170362628
#>  [49] 0.158880757 0.148172726 0.138186380 0.128873081 0.120187466 0.112087233
#>  [55] 0.104532929 0.097487759 0.090917411 0.084789882 0.079075327 0.073745915
#>  [61] 0.068775687 0.064140436 0.059817585 0.055786080 0.052026285 0.048519888
#>  [67] 0.045249810 0.042200124 0.039355977 0.036703516 0.034229822 0.031922846
#>  [73] 0.029771353 0.027764864 0.025893605 0.024148462 0.022520937 0.021003101
#>  [79] 0.019587562 0.018267426 0.017036263 0.015888076 0.014817273 0.013818638
#>  [85] 0.012887308 0.012018747 0.011208723 0.010453293 0.009748776 0.009091741
#>  [91] 0.008478988 0.007907533 0.007374591 0.006877569 0.006414044 0.005981759
#>  [97] 0.005578608 0.005202629 0.004851989 0.004524981
#> 
#> $fit
#> $beta
#>              4.5250     4.2200     3.9356    3.6704    3.4230    3.1923
#> (Intercept) 18.1619 19.7001939 21.1348075 22.472733 23.720487 24.884146
#> am           0.0000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> carb         0.0000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> cyl          0.0000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> disp         0.0000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> drat         0.0000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> gear         0.0000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> hp           0.0000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> qsec         0.0000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> vs           0.0000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> wt           0.0000 -0.4515589 -0.8726842 -1.265427 -1.631701 -1.973288
#>                2.9771   2.7765    2.5894    2.4148    2.2521    2.1003
#> (Intercept) 25.969379 26.98147 27.925350 28.805616 29.626555 30.392165
#> am           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.291854 -2.58895 -2.866022 -3.124421 -3.365404 -3.590146
#>                1.9588   1.8267    1.7036    1.5888   1.4817   1.3819     1.2887
#> (Intercept) 31.106175 31.77206 32.393074 32.972230 33.37818 33.37818 31.6135996
#> am           0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.0000000
#> carb         0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.0000000
#> cyl          0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.0000000
#> disp         0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.0000000
#> drat         0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.0000000
#> gear         0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.0000000
#> hp           0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.0000000
#> qsec         0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.1009346
#> vs           0.000000  0.00000  0.000000  0.000000  0.00000  0.00000  0.0000000
#> wt          -3.799741 -3.99521 -4.177505 -4.347514 -4.46668 -4.46668 -4.4560543
#>                 1.2019     1.1209     1.0453     0.9749     0.9092     0.8479
#> (Intercept) 29.8603470 28.2252577 26.7003679 25.2782505 23.9519792 22.7150940
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.2012212  0.2947487  0.3819729  0.4633184  0.5391815  0.6099317
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -4.4454970 -4.4356513 -4.4264691 -4.4179058 -4.4099196 -4.4024716
#>                 0.7908     0.7375     0.6878     0.6414     0.5982     0.5579
#> (Intercept) 21.5615708 20.4857912 19.4825156 18.5468573 17.6742593 16.8604715
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.6759136  0.7374485  0.7948361  0.8483561  0.8982689  0.9448178
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -4.3955256 -4.3890478 -4.3830065 -4.3773724 -4.3721181 -4.3672178
#>                 0.5203     0.4852    0.4525    0.4220    0.3936    0.3670
#> (Intercept) 16.1015303 13.7270484 10.312226  7.556727  4.990645  4.609220
#> am           0.0000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.0000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.0000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.0000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.0000000  0.3373772  0.929237  1.484788  2.002161  2.079039
#> gear         0.0000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.0000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.9882295  1.0276430  1.053844  1.052708  1.051646  1.051495
#> vs           0.0000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt          -4.3626478 -4.2173498 -3.966974 -3.734668 -3.518325 -3.486178
#>                0.3423     0.3192     0.2977     0.2776     0.2589     0.2415
#> (Intercept)  4.607271 -10.445418 -10.450754 -10.454679 -10.457565 -10.459686
#> am           0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> carb         0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> cyl          0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> disp         0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> drat         2.079406   3.266239   3.266791   3.267197   3.267495   3.267715
#> gear         0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> hp           0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> qsec         1.051502   1.780781   1.780966   1.781102   1.781202   1.781276
#> vs           0.000000  -3.556829  -3.557475  -3.557949  -3.558297  -3.558552
#> wt          -3.486023  -3.728505  -3.728403  -3.728328  -3.728272  -3.728231
#>                 0.2252     0.2100     0.1959     0.1827     0.1704     0.1589
#> (Intercept) -10.461244 -10.462388 -10.463228 -12.257210 -12.259310 -12.260332
#> am            0.000000   0.000000   0.000000   1.238049   1.237675   1.237340
#> carb          0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> cyl           0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> disp          0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> drat          3.267877   3.267995   3.268082   2.781104   2.781637   2.781992
#> gear          0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> hp            0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> qsec          1.781330   1.781369   1.781398   1.920535   1.920540   1.920528
#> vs           -3.558739  -3.558877  -3.558978  -3.805035  -3.805076  -3.805067
#> wt           -3.728201  -3.728178  -3.728162  -3.529086  -3.529006  -3.528977
#>                 0.1482     0.1382     0.1289        0.1202        0.1121
#> (Intercept) -12.260703 -12.260709 -12.260531 -16.551908627 -16.554477929
#> am            1.237063   1.236845   1.236681   1.830175283   1.830869342
#> carb          0.000000   0.000000   0.000000   0.000000000   0.000000000
#> cyl           0.000000   0.000000   0.000000   0.000000000   0.000000000
#> disp          0.000000   0.000000   0.000000   0.009862997   0.009866201
#> drat          2.782220   2.782360   2.782439   2.784609363   2.784365552
#> gear          0.000000   0.000000   0.000000   0.000000000   0.000000000
#> hp            0.000000   0.000000   0.000000   0.000000000   0.000000000
#> qsec          1.920509   1.920487   1.920467   2.165997816   2.166185820
#> vs           -3.805035  -3.804995  -3.804954  -3.644522646  -3.644788040
#> wt           -3.528976  -3.528989  -3.529007  -4.369790407  -4.370047202
#>                    0.1045      0.0975      0.0909      0.0848      0.0791
#> (Intercept) -16.556660303 -6.41621145 -6.41709550 -6.41794748 -6.41876853
#> am            1.831453975  0.00000000  0.00000000  0.00000000  0.00000000
#> carb          0.000000000  0.54681958  0.54693885  0.54705379  0.54716457
#> cyl           0.000000000 -1.26908177 -1.26922714 -1.26936723 -1.26950223
#> disp          0.009868893  0.02440478  0.02440864  0.02441236  0.02441594
#> drat          2.784162252  1.99624381  1.99607522  1.99591275  1.99575617
#> gear          0.000000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp            0.000000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec          2.166344728  2.24493398  2.24508052  2.24522173  2.24535782
#> vs           -3.645013979 -4.15998813 -4.16017318 -4.16035151 -4.16052337
#> wt           -4.370262455 -5.79733877 -5.79775538 -5.79815688 -5.79854381
#>                  0.0737      0.0688      0.0641      0.0598      0.0558
#> (Intercept) -6.41955979 -6.42032234 -6.42105722 -6.42176543 -6.42244794
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.54727132  0.54737420  0.54747334  0.54756889  0.54766097
#> cyl         -1.26963234 -1.26975772 -1.26987855 -1.26999500 -1.27010723
#> disp         0.02441939  0.02442272  0.02442592  0.02442902  0.02443199
#> drat         1.99560527  1.99545985  1.99531971  1.99518465  1.99505449
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         2.24548898  2.24561537  2.24573718  2.24585457  2.24596769
#> vs          -4.16068900 -4.16084862 -4.16100244 -4.16115068 -4.16129354
#> wt          -5.79891669 -5.79927605 -5.79962237 -5.79995611 -5.80027775
#>                  0.0520      0.0485      0.0452      0.0422      0.0394
#> (Intercept) -6.42310568 -6.42373955 -6.42435042 -6.42493912 -6.42550646
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.54774971  0.54783522  0.54791764  0.54799706  0.54807361
#> cyl         -1.27021538 -1.27031960 -1.27042005 -1.27051685 -1.27061013
#> disp         0.02443486  0.02443763  0.02444029  0.02444286  0.02444534
#> drat         1.99492906  1.99480818  1.99469168  1.99457942  1.99447122
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         2.24607671  2.24618178  2.24628303  2.24638061  2.24647465
#> vs          -4.16143122 -4.16156390 -4.16169177 -4.16181500 -4.16193375
#> wt          -5.80058771 -5.80088643 -5.80117431 -5.80145174 -5.80171910
#>                  0.0367      0.0342      0.0319      0.0298      0.0278
#> (Intercept) -6.42605322 -6.42658013 -6.42708792 -6.42757728 -6.42804889
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.54814737  0.54821846  0.54828697  0.54835299  0.54841661
#> cyl         -1.27070003 -1.27078667 -1.27087017 -1.27095063 -1.27102818
#> disp         0.02444773  0.02445002  0.02445224  0.02445438  0.02445643
#> drat         1.99436696  1.99426647  1.99416963  1.99407631  1.99398637
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         2.24656528  2.24665261  2.24673678  2.24681789  2.24689606
#> vs          -4.16204820 -4.16215849 -4.16226478 -4.16236721 -4.16246593
#> wt          -5.80197676 -5.80222507 -5.80246437 -5.80269499 -5.80291724
#>                  0.0259      0.0241      0.0225      0.0210      0.0196
#> (Intercept) -6.42850338 -6.42894138 -6.42936349 -6.42977027 -6.43016230
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.54847793  0.54853702  0.54859397  0.54864885  0.54870174
#> cyl         -1.27110291 -1.27117493 -1.27124433 -1.27131122 -1.27137568
#> disp         0.02445842  0.02446033  0.02446217  0.02446394  0.02446566
#> drat         1.99389970  1.99381617  1.99373568  1.99365810  1.99358334
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         2.24697140  2.24704399  2.24711396  2.24718139  2.24724636
#> vs          -4.16256106 -4.16265274 -4.16274110 -4.16282625 -4.16290831
#> wt          -5.80313142 -5.80333783 -5.80353675 -5.80372845 -5.80391319
#>                 0.0183      0.0170     0.0159     0.0148      0.0138
#> (Intercept) -6.4305401 -6.45719424 -6.4578094 -6.4583646 -6.45886500
#> am           0.0000000  0.41114866  0.4113872  0.4116176  0.41183959
#> carb         0.5487527  0.55333877  0.5531996  0.5530661  0.55293800
#> cyl         -1.2714378 -1.21561561 -1.2153900 -1.2151748 -1.21496937
#> disp         0.0244673  0.02389514  0.0238914  0.0238878  0.02388433
#> drat         1.9935113  2.03900973  2.0391504  2.0392857  2.03941603
#> gear         0.0000000 -0.21285621 -0.2128199 -0.2127878 -0.21275947
#> hp           0.0000000  0.00000000  0.0000000  0.0000000  0.00000000
#> qsec         2.2473090  2.24906803  2.2489722  2.2488789  2.24878820
#> vs          -4.1629874 -4.17257155 -4.1723684 -4.1721745 -4.17198948
#> wt          -5.8040912 -5.74476655 -5.7443557 -5.7439591 -5.74357624
#>                  0.0129      0.0120      0.0112      0.0105      0.0097
#> (Intercept) -6.45931561 -6.45972108 -6.46008579 -6.46041382 -6.46070894
#> am           0.41205320  0.41225847  0.41245550  0.41264442  0.41282542
#> carb         0.55281507  0.55269704  0.55258367  0.55247472  0.55236997
#> cyl         -1.21477321 -1.21458584 -1.21440675 -1.21423550 -1.21407164
#> disp         0.02388099  0.02387776  0.02387464  0.02387163  0.02386872
#> drat         2.03954163  2.03966284  2.03977994  2.03989315  2.04000268
#> gear        -0.21273445 -0.21271234 -0.21269278 -0.21267544 -0.21266001
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         2.24869994  2.24861410  2.24853063  2.24844948  2.24837061
#> vs          -4.17181282 -4.17164415 -4.17148308 -4.17132921 -4.17118215
#> wt          -5.74320649 -5.74284932 -5.74250423 -5.74217072 -5.74184836
#>                  0.0091     0.0085      0.0079      0.0074      0.0069
#> (Intercept) -6.46097464 -6.4612141 -6.46143028 -6.46162579 -6.46180304
#> am           0.41299871  0.4131645  0.41332314  0.41347481  0.41361982
#> carb         0.55226922  0.5521723  0.55207894  0.55198906  0.55190246
#> cyl         -1.21391476 -1.2137645 -1.21362038 -1.21348217 -1.21334950
#> disp         0.02386591  0.0238632  0.02386057  0.02385804  0.02385558
#> drat         2.04010871  2.0402114  2.04031092  2.04040737  2.04050087
#> gear        -0.21264623 -0.2126338 -0.21262267 -0.21261250 -0.21260318
#> hp           0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> qsec         2.24829396  2.2482195  2.24814719  2.24807697  2.24800882
#> vs          -4.17104153 -4.1709070 -4.17077820 -4.17065484 -4.17053659
#> wt          -5.74153673 -5.7412354 -5.74094404 -5.74066225 -5.74038972
#>                  0.0064      0.0060        0.0056       0.0052      0.0049
#> (Intercept) -6.46196421 -6.46256472 -6.458918e+00 -5.846757933 -5.84313639
#> am           0.41375845  0.41399547  4.141178e-01  0.440809954  0.44117615
#> carb         0.55181899  0.55167669  5.515973e-01  0.557663558  0.55775210
#> cyl         -1.21322209 -1.21297705 -1.212860e+00 -1.186378415 -1.18628523
#> disp         0.02385321  0.02384872  2.384659e-02  0.023862054  0.02386452
#> drat         2.04059154  2.04077207  2.040858e+00  2.033045086  2.03281114
#> gear        -0.21259458 -0.21257759 -2.125666e-01 -0.208430256 -0.20850428
#> hp           0.00000000  0.00000000 -5.242059e-06 -0.001329689 -0.00133618
#> qsec         2.24794267  2.24783139  2.247568e+00  2.202844022  2.20265763
#> vs          -4.17042317 -4.17019229 -4.170248e+00 -4.072921942 -4.07227146
#> wt          -5.74012611 -5.73964063 -5.739227e+00 -5.694031627 -5.69406442
#>                   0.0045
#> (Intercept) -5.839695110
#> am           0.441519054
#> carb         0.557838883
#> cyl         -1.186203934
#> disp         0.023866935
#> drat         2.032586224
#> gear        -0.208577831
#> hp          -0.001342264
#> qsec         2.202484489
#> vs          -4.071664011
#> wt          -5.694105884
#> 
#> $iter
#>   [1]   0   4   3   3   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   4   3   3   3   3   3   3   3   3   3   3   3   3   3  30  30  30  29
#>  [37]   8   1  34   1   1   1   1   1   1   1  58   1   1   1   1   1  52   1
#>  [55]   1 207   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1 111   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   2   1  50   1   1
#> 
#> $lambda
#>   [1] 4.524980958 4.220012386 3.935597675 3.670351564 3.422982153 3.192284612
#>   [7] 2.977135313 2.776486357 2.589360469 2.414846240 2.252093686 2.100310109
#>  [13] 1.958756237 1.826742622 1.703626283 1.588807574 1.481727261 1.381863803
#>  [19] 1.288730807 1.201874664 1.120872334 1.045329290 0.974877595 0.909174108
#>  [25] 0.847898816 0.790753273 0.737459149 0.687756871 0.641404360 0.598175852
#>  [31] 0.557860802 0.520262851 0.485198876 0.452498096 0.422001239 0.393559768
#>  [37] 0.367035156 0.342298215 0.319228461 0.297713531 0.277648636 0.258936047
#>  [43] 0.241484624 0.225209369 0.210031011 0.195875624 0.182674262 0.170362628
#>  [49] 0.158880757 0.148172726 0.138186380 0.128873081 0.120187466 0.112087233
#>  [55] 0.104532929 0.097487759 0.090917411 0.084789882 0.079075327 0.073745915
#>  [61] 0.068775687 0.064140436 0.059817585 0.055786080 0.052026285 0.048519888
#>  [67] 0.045249810 0.042200124 0.039355977 0.036703516 0.034229822 0.031922846
#>  [73] 0.029771353 0.027764864 0.025893605 0.024148462 0.022520937 0.021003101
#>  [79] 0.019587562 0.018267426 0.017036263 0.015888076 0.014817273 0.013818638
#>  [85] 0.012887308 0.012018747 0.011208723 0.010453293 0.009748776 0.009091741
#>  [91] 0.008478988 0.007907533 0.007374591 0.006877569 0.006414044 0.005981759
#>  [97] 0.005578608 0.005202629 0.004851989 0.004524981
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 32
#> 
#> $loss
#>   [1] 550.60952 468.06550 399.00515 341.48807 293.83900 254.61247 222.56207
#>   [8] 196.61380 175.84295 159.45399 146.76319 137.18350 130.21141 125.41563
#>  [15] 122.42722 120.93106 120.62502 120.62502 113.33622 106.81899 101.39206
#>  [22]  96.89713  93.19763  90.17582  87.73022  85.77348  84.23044  83.03651
#>  [29]  82.13623  81.48206  81.03324  80.75493  76.93191  72.01023  69.23796
#>  [36]  68.24292  68.22566  68.22566  57.63109  57.63108  57.63108  57.63108
#>  [43]  57.63108  57.63108  57.63108  57.63108  55.26499  55.26499  55.26499
#>  [50]  55.26498  55.26498  55.26498  52.06071  52.06070  52.06070  46.79838
#>  [57]  46.79836  46.79835  46.79834  46.79833  46.79832  46.79832  46.79831
#>  [64]  46.79830  46.79829  46.79829  46.79828  46.79828  46.79827  46.79827
#>  [71]  46.79826  46.79826  46.79826  46.79825  46.79825  46.79825  46.79824
#>  [78]  46.79824  46.79824  46.79824  46.67664  46.67663  46.67662  46.67660
#>  [85]  46.67659  46.67658  46.67657  46.67657  46.67656  46.67655  46.67654
#>  [92]  46.67654  46.67653  46.67653  46.67652  46.67651  46.67642  46.66425
#>  [99]  46.66424  46.66423
#> 
#> $linear.predictors
#>     4.5250   4.2200   3.9356   3.6704   3.4230   3.1923   2.9771   2.7765
#> 1  18.1619 18.51711 18.84837 19.15731 19.44543 19.71413 19.96472 20.19842
#> 2  18.1619 18.65258 19.11018 19.53694 19.93494 20.30612 20.65228 20.97511
#> 3  18.1619 18.14683 18.13277 18.11966 18.10744 18.09603 18.08540 18.07548
#> 4  18.1619 18.08813 18.01932 17.95516 17.89532 17.83951 17.78746 17.73892
#> 5  18.1619 18.14683 18.13277 18.11966 18.10744 18.09603 18.08540 18.07548
#> 6  18.1619 18.01588 17.87970 17.75269 17.63424 17.52378 17.42076 17.32469
#> 7  18.1619 17.99330 17.83606 17.68942 17.55266 17.42512 17.30617 17.19524
#> 8  18.1619 17.32951 16.55322 15.82924 15.15406 14.52438 13.93714 13.38948
#> 9  18.1619 17.25094 16.40137 15.60906 14.87014 14.18103 13.53836 12.93901
#> 10 18.1619 17.28661 16.47031 15.70902 14.99905 14.33692 13.71942 13.14353
#> 11 18.1619 18.97093 19.72542 20.42907 21.08529 21.69729 22.26803 22.80032
#> 12 18.1619 18.11071 18.06296 18.01843 17.97690 17.93817 17.90205 17.86837
#> 13 18.1619 18.14909 18.13714 18.12599 18.11560 18.10590 18.09686 18.08843
#> 14 18.1619 17.96621 17.78370 17.61349 17.45476 17.30672 17.16866 17.03990
#> 15 18.1619 17.96395 17.77934 17.60717 17.44660 17.29685 17.15720 17.02696
#> 16 18.1619 18.82643 19.44616 20.02413 20.56315 21.06583 21.53464 21.97185
#> 17 18.1619 18.73386 19.26726 19.76472 20.22865 20.66131 21.06481 21.44112
#> 18 18.1619 18.26875 18.36840 18.46133 18.54800 18.62882 18.70420 18.77450
#> 19 18.1619 18.44938 18.71747 18.96750 19.20068 19.41814 19.62094 19.81008
#> 20 18.1619 18.08813 18.01932 17.95516 17.89532 17.83951 17.78746 17.73892
#> 21 18.1619 18.44486 18.70875 18.95485 19.18436 19.39840 19.59802 19.78419
#>      2.5894   2.4148   2.2521   2.1003   1.9588   1.8267    1.7036    1.5888
#> 1  20.41637 20.61963 20.80920 20.98598 21.15085 21.30461 21.448010 21.581742
#> 2  21.27618 21.55696 21.81882 22.06303 22.29078 22.50318 22.701262 22.885997
#> 3  18.06623 18.05761 18.04956 18.04206 18.03507 18.02854 18.022456 18.016781
#> 4  17.69365 17.65143 17.61206 17.57534 17.54110 17.50916 17.479380 17.451604
#> 5  18.06623 18.05761 18.04956 18.04206 18.03507 18.02854 18.022456 18.016781
#> 6  17.23509 17.15153 17.07360 17.00092 16.93314 16.86993 16.810979 16.756001
#> 7  17.09179 16.99531 16.90533 16.82141 16.74315 16.67017 16.602104 16.538626
#> 8  12.87873 12.40241 11.95818 11.54390 11.15753 10.79721 10.461171 10.147780
#> 9  12.38005 11.85876 11.37260 10.91921 10.49638 10.10204  9.734285  9.391312
#> 10 12.60646 12.10559 11.63847 11.20283 10.79656 10.41767 10.064308  9.734766
#> 11 23.29672 23.75968 24.19143 24.59408 24.96959 25.31980 25.646403 25.950994
#> 12 17.83695 17.80765 17.78033 17.75485 17.73109 17.70892 17.688255 17.668979
#> 13 18.08056 18.07323 18.06639 18.06001 18.05406 18.04852 18.043343 18.038518
#> 14 16.91983 16.80784 16.70340 16.60600 16.51517 16.43046 16.351454 16.277775
#> 15 16.90550 16.79222 16.68658 16.58805 16.49617 16.41048 16.330566 16.256037
#> 16 22.37960 22.75986 23.11450 23.44523 23.75368 24.04133 24.309601 24.559790
#> 17 21.79206 22.11936 22.42459 22.70925 22.97473 23.22231 23.453213 23.668549
#> 18 18.84006 18.90120 18.95822 19.01140 19.06100 19.10725 19.150382 19.190609
#> 19 19.98647 20.15097 20.30439 20.44746 20.58089 20.70533 20.821384 20.929615
#> 20 17.69365 17.65143 17.61206 17.57534 17.54110 17.50916 17.479380 17.451604
#> 21 19.95781 20.11973 20.27073 20.41156 20.54289 20.66538 20.779609 20.886140
#>       1.4817    1.3819    1.2887    1.2019    1.1209    1.0453    0.9749
#> 1  21.675480 21.675480 21.600121 21.525245 21.455416 21.390292 21.329558
#> 2  23.015484 23.015484 23.153946 23.291519 23.419821 23.539475 23.651065
#> 3  18.012803 18.012803 18.002679 17.992621 17.983241 17.974493 17.966334
#> 4  17.432134 17.432134 17.304290 17.177266 17.058803 16.948324 16.845291
#> 5  18.012803 18.012803 18.192436 18.370917 18.537368 18.692602 18.837373
#> 6  16.717466 16.717466 16.768966 16.820135 16.867856 16.912361 16.953866
#> 7  16.494132 16.494132 16.586537 16.678349 16.763973 16.843827 16.918298
#> 8   9.928112  9.928112 10.034118 10.139444 10.237671 10.329277 10.414710
#> 9   9.150910  9.150910  9.242615  9.333732  9.418708  9.497956  9.571864
#> 10  9.503778  9.503778  9.554270  9.604438  9.651225  9.694858  9.735551
#> 11 26.164493 26.164493 26.286380 26.407485 26.520427 26.625758 26.723990
#> 12 17.655468 17.655468 17.631055 17.606798 17.584176 17.563079 17.543404
#> 13 18.035136 18.035136 18.053221 18.071191 18.087949 18.103577 18.118153
#> 14 16.226131 16.226131 16.057753 15.890456 15.734435 15.588929 15.453229
#> 15 16.203797 16.203797 16.201005 16.198232 16.195644 16.193232 16.190982
#> 16 24.735156 24.735156 24.898798 25.061390 25.213024 25.354438 25.486321
#> 17 23.819486 23.819486 23.763251 23.707377 23.655268 23.606671 23.561350
#> 18 19.218806 19.218806 18.951459 18.685828 18.438100 18.207068 17.991606
#> 19 21.005478 21.005478 20.834815 20.665248 20.507109 20.359628 20.222087
#> 20 17.432134 17.432134 17.179131 16.927751 16.693314 16.474677 16.270776
#> 21 20.960811 20.960811 21.103152 21.244579 21.376474 21.499479 21.614195
#>       0.9092    0.8479    0.7908    0.7375    0.6878    0.6414    0.5982
#> 1  21.272918 21.220094 21.170831 21.124888 21.082041 21.042082 21.004817
#> 2  23.755134 23.852189 23.942703 24.027117 24.105841 24.179260 24.247730
#> 3  17.958725 17.951629 17.945012 17.938840 17.933084 17.927716 17.922710
#> 4  16.749202 16.659589 16.576015 16.498075 16.425387 16.357598 16.294378
#> 5  18.972387 19.098301 19.215729 19.325243 19.427376 19.522626 19.611456
#> 6  16.992574 17.028673 17.062339 17.093736 17.123017 17.150325 17.175792
#> 7  16.987751 17.052522 17.112928 17.169263 17.221801 17.270799 17.316494
#> 8  10.494385 10.568690 10.637987 10.702614 10.762885 10.819094 10.871515
#> 9   9.640790  9.705071  9.765020  9.820928  9.873068  9.921694  9.967043
#> 10  9.773501  9.808894  9.841901  9.872683  9.901391  9.928164  9.953133
#> 11 26.815601 26.901038 26.980716 27.055025 27.124325 27.188955 27.249229
#> 12 17.525054 17.507942 17.491983 17.477099 17.463218 17.450273 17.438201
#> 13 18.131746 18.144423 18.156245 18.167271 18.177553 18.187143 18.196086
#> 14 15.326675 15.208651 15.098581 14.995929 14.900195 14.810914 14.727650
#> 15 16.188883 16.186926 16.185101 16.183399 16.181812 16.180331 16.178951
#> 16 25.609315 25.724021 25.830995 25.930760 26.023801 26.110571 26.191494
#> 17 23.519083 23.479664 23.442903 23.408619 23.376645 23.346827 23.319018
#> 18 17.790666 17.603269 17.428501 17.265513 17.113509 16.971750 16.839545
#> 19 20.093815 19.974189 19.862625 19.758580 19.661548 19.571055 19.486661
#> 20 16.080616 15.903273 15.737882 15.583638 15.439790 15.305636 15.180524
#> 21 21.721179 21.820953 21.914002 22.000780 22.081710 22.157185 22.227573
#>       0.5579    0.5203   0.4852   0.4525   0.4220   0.3936   0.3670   0.3423
#> 1  20.970062 20.937650 20.90837 20.88906 20.89014 20.89115 20.89130 20.89129
#> 2  24.311586 24.371138 24.36614 24.29845 24.19962 24.10758 24.09391 24.09386
#> 3  17.918042 17.913688 17.77259 17.52936 17.30364 17.09343 17.06219 17.06204
#> 4  16.235418 16.180433 16.03196 15.82587 15.66502 15.51523 15.49297 15.49285
#> 5  19.694299 19.771559 19.96434 20.22610 20.42602 20.61219 20.63986 20.64001
#> 6  17.199543 17.221693 17.11860 16.91583 16.71237 16.52290 16.49474 16.49461
#> 7  17.359109 17.398852 17.31879 17.13902 16.94672 16.76764 16.74103 16.74091
#> 8  10.920403 10.965996 11.05150 11.15640 11.22784 11.29436 11.30426 11.30431
#> 9  10.009336 10.048778 10.17687 10.36258 10.51351 10.65406 10.67496 10.67506
#> 10  9.976419  9.998136 10.17658 10.46815 10.72896 10.97185 11.00794 11.00812
#> 11 27.305441 27.357864 27.61125 28.00390 28.34139 28.65569 28.70240 28.70263
#> 12 17.426942 17.416442 17.14948 16.69153 16.26789 15.87337 15.81475 15.81447
#> 13 18.204427 18.212205 18.08141 17.84427 17.61707 17.40548 17.37404 17.37389
#> 14 14.649998 14.577579 14.62682 14.78484 14.97609 15.15420 15.18066 15.18077
#> 15 16.177663 16.176462 16.07177 15.88931 15.71874 15.55990 15.53630 15.53619
#> 16 26.266962 26.337344 26.36543 26.34508 26.28426 26.22761 26.21921 26.21918
#> 17 23.293083 23.268897 23.35814 23.53862 23.72237 23.89349 23.91891 23.91903
#> 18 16.716250 16.601264 16.68260 16.93904 17.24790 17.53554 17.57826 17.57845
#> 19 19.407955 19.334553 19.19476 19.02213 18.90360 18.79322 18.77681 18.77671
#> 20 15.063844 14.955028 14.86901 14.82575 14.84965 14.87190 14.87519 14.87519
#> 21 22.293218 22.354438 22.50360 22.70471 22.85719 22.99920 23.02031 23.02042
#>      0.3192   0.2977   0.2776   0.2589   0.2415   0.2252   0.2100   0.1959
#> 1  21.83589 21.83602 21.83612 21.83619 21.83624 21.83628 21.83630 21.83633
#> 2  23.06298 23.06280 23.06268 23.06258 23.06251 23.06246 23.06243 23.06240
#> 3  17.32607 17.32598 17.32591 17.32586 17.32582 17.32579 17.32577 17.32575
#> 4  14.93602 14.93575 14.93555 14.93541 14.93530 14.93523 14.93517 14.93513
#> 5  19.63211 19.63215 19.63217 19.63219 19.63220 19.63221 19.63222 19.63223
#> 6  17.01636 17.01636 17.01636 17.01635 17.01635 17.01635 17.01635 17.01635
#> 7  17.54224 17.54232 17.54238 17.54242 17.54245 17.54247 17.54249 17.54250
#> 8  11.56845 11.56860 11.56871 11.56879 11.56885 11.56889 11.56892 11.56895
#> 9  10.86340 10.86358 10.86371 10.86380 10.86387 10.86392 10.86396 10.86399
#> 10 11.19688 11.19710 11.19726 11.19738 11.19747 11.19753 11.19758 11.19761
#> 11 29.05884 29.05917 29.05942 29.05960 29.05973 29.05983 29.05990 29.05996
#> 12 15.48684 15.48651 15.48627 15.48609 15.48596 15.48586 15.48579 15.48574
#> 13 17.84333 17.84329 17.84326 17.84323 17.84322 17.84320 17.84319 17.84319
#> 14 14.86203 14.86200 14.86197 14.86196 14.86195 14.86194 14.86193 14.86193
#> 15 15.64081 15.64073 15.64066 15.64062 15.64058 15.64056 15.64054 15.64052
#> 16 25.76611 25.76608 25.76606 25.76604 25.76603 25.76602 25.76601 25.76601
#> 17 25.78406 25.78448 25.78479 25.78502 25.78518 25.78531 25.78540 25.78546
#> 18 17.34007 17.34008 17.34008 17.34008 17.34008 17.34008 17.34008 17.34008
#> 19 18.65251 18.65233 18.65219 18.65209 18.65202 18.65196 18.65192 18.65190
#> 20 13.80571 13.80539 13.80516 13.80499 13.80487 13.80478 13.80471 13.80466
#> 21 22.17928 22.17929 22.17930 22.17931 22.17932 22.17932 22.17933 22.17933
#>      0.1827   0.1704   0.1589   0.1482   0.1382   0.1289   0.1202   0.1121
#> 1  22.19294 22.19284 22.19275 22.19267 22.19261 22.19256 21.91980 21.91990
#> 2  23.43672 23.43655 23.43641 23.43631 23.43624 23.43619 23.59100 23.59117
#> 3  17.05071 17.05066 17.05065 17.05067 17.05070 17.05073 17.60349 17.60363
#> 4  14.49257 14.49255 14.49258 14.49264 14.49270 14.49275 14.64662 14.64648
#> 5  18.99773 18.99806 18.99831 18.99850 18.99864 18.99873 18.27756 18.27697
#> 6  16.91870 16.91863 16.91860 16.91858 16.91858 16.91859 16.53930 16.53922
#> 7  17.51046 17.51040 17.51036 17.51034 17.51033 17.51033 17.18721 17.18719
#> 8  11.89494 11.89492 11.89487 11.89482 11.89478 11.89474 12.26557 12.26583
#> 9  11.16827 11.16830 11.16829 11.16826 11.16822 11.16818 11.23524 11.23537
#> 10 11.31850 11.31865 11.31872 11.31875 11.31876 11.31875 11.15725 11.15721
#> 11 28.75547 28.75582 28.75605 28.75619 28.75628 28.75633 29.16557 29.16553
#> 12 15.39567 15.39542 15.39528 15.39521 15.39518 15.39518 15.42877 15.42881
#> 13 17.60611 17.60606 17.60605 17.60606 17.60608 17.60610 17.67949 17.67950
#> 14 14.16006 14.16034 14.16057 14.16075 14.16089 14.16099 13.88476 13.88432
#> 15 15.48437 15.48431 15.48429 15.48430 15.48431 15.48432 16.09831 16.09848
#> 16 25.99203 25.99195 25.99188 25.99183 25.99179 25.99176 26.25594 26.25612
#> 17 25.82181 25.82196 25.82204 25.82208 25.82209 25.82209 25.62142 25.62144
#> 18 17.37765 17.37776 17.37782 17.37785 17.37787 17.37787 18.04596 18.04610
#> 19 19.04115 19.04092 19.04074 19.04062 19.04053 19.04048 18.25733 18.25724
#> 20 14.26692 14.26669 14.26653 14.26640 14.26632 14.26626 14.12796 14.12802
#> 21 22.51723 22.51723 22.51720 22.51716 22.51712 22.51708 22.41145 22.41148
#>      0.1045   0.0975   0.0909   0.0848   0.0791   0.0737   0.0688   0.0641
#> 1  21.92000 21.60928 21.60928 21.60928 21.60928 21.60928 21.60928 21.60928
#> 2  23.59131 23.54394 23.54394 23.54394 23.54393 23.54393 23.54393 23.54393
#> 3  17.60374 17.86459 17.86470 17.86481 17.86491 17.86501 17.86511 17.86520
#> 4  14.64637 15.67533 15.67544 15.67555 15.67565 15.67575 15.67585 15.67594
#> 5  18.27648 18.39851 18.39837 18.39823 18.39810 18.39797 18.39785 18.39773
#> 6  16.53915 15.81766 15.81755 15.81743 15.81732 15.81722 15.81712 15.81702
#> 7  17.18717 16.42577 16.42569 16.42561 16.42554 16.42547 16.42540 16.42533
#> 8  12.26605 12.91435 12.91455 12.91475 12.91494 12.91512 12.91530 12.91547
#> 9  11.23548 11.39330 11.39335 11.39340 11.39345 11.39349 11.39354 11.39358
#> 10 11.15717 10.92436 10.92427 10.92418 10.92409 10.92401 10.92393 10.92386
#> 11 29.16551 29.34351 29.34360 29.34369 29.34377 29.34385 29.34393 29.34400
#> 12 15.42885 15.26053 15.26049 15.26045 15.26041 15.26037 15.26034 15.26030
#> 13 17.67951 17.15549 17.15543 17.15537 17.15531 17.15526 17.15520 17.15515
#> 14 13.88394 13.93873 13.93854 13.93835 13.93818 13.93801 13.93784 13.93768
#> 15 16.09864 16.42047 16.42059 16.42069 16.42080 16.42090 16.42099 16.42109
#> 16 26.25626 26.17835 26.17840 26.17844 26.17849 26.17853 26.17857 26.17862
#> 17 25.62146 26.46445 26.46449 26.46454 26.46458 26.46462 26.46466 26.46470
#> 18 18.04621 16.78262 16.78249 16.78238 16.78226 16.78215 16.78205 16.78194
#> 19 18.25717 18.75316 18.75319 18.75321 18.75323 18.75326 18.75328 18.75330
#> 20 14.12806 14.29777 14.29789 14.29801 14.29813 14.29824 14.29834 14.29845
#> 21 22.41150 22.23782 22.23775 22.23768 22.23762 22.23755 22.23749 22.23744
#>      0.0598   0.0558   0.0520   0.0485   0.0452   0.0422   0.0394   0.0367
#> 1  21.60928 21.60928 21.60929 21.60929 21.60929 21.60929 21.60929 21.60929
#> 2  23.54392 23.54392 23.54392 23.54392 23.54391 23.54391 23.54391 23.54391
#> 3  17.86529 17.86537 17.86545 17.86553 17.86561 17.86568 17.86575 17.86582
#> 4  15.67603 15.67612 15.67620 15.67628 15.67636 15.67643 15.67650 15.67657
#> 5  18.39762 18.39751 18.39740 18.39730 18.39720 18.39711 18.39702 18.39693
#> 6  15.81692 15.81683 15.81675 15.81666 15.81658 15.81650 15.81643 15.81635
#> 7  16.42527 16.42521 16.42515 16.42509 16.42503 16.42498 16.42493 16.42488
#> 8  12.91563 12.91579 12.91594 12.91609 12.91623 12.91637 12.91650 12.91662
#> 9  11.39362 11.39366 11.39369 11.39373 11.39376 11.39380 11.39383 11.39386
#> 10 10.92378 10.92371 10.92365 10.92358 10.92352 10.92346 10.92340 10.92334
#> 11 29.34408 29.34415 29.34421 29.34428 29.34434 29.34440 29.34446 29.34451
#> 12 15.26027 15.26024 15.26021 15.26018 15.26015 15.26012 15.26009 15.26007
#> 13 17.15510 17.15505 17.15500 17.15496 17.15492 17.15487 17.15483 17.15479
#> 14 13.93753 13.93738 13.93724 13.93711 13.93698 13.93685 13.93673 13.93661
#> 15 16.42118 16.42126 16.42135 16.42143 16.42150 16.42158 16.42165 16.42172
#> 16 26.17865 26.17869 26.17873 26.17876 26.17880 26.17883 26.17886 26.17889
#> 17 26.46474 26.46477 26.46481 26.46484 26.46487 26.46490 26.46493 26.46496
#> 18 16.78184 16.78175 16.78166 16.78157 16.78148 16.78140 16.78132 16.78125
#> 19 18.75332 18.75334 18.75336 18.75338 18.75340 18.75341 18.75343 18.75345
#> 20 14.29855 14.29864 14.29873 14.29882 14.29891 14.29899 14.29907 14.29915
#> 21 22.23738 22.23732 22.23727 22.23722 22.23717 22.23712 22.23708 22.23703
#>      0.0342   0.0319   0.0298   0.0278   0.0259   0.0241   0.0225   0.0210
#> 1  21.60929 21.60929 21.60929 21.60929 21.60929 21.60929 21.60929 21.60929
#> 2  23.54390 23.54390 23.54390 23.54390 23.54390 23.54390 23.54389 23.54389
#> 3  17.86588 17.86595 17.86601 17.86607 17.86612 17.86618 17.86623 17.86628
#> 4  15.67664 15.67670 15.67676 15.67682 15.67688 15.67694 15.67699 15.67704
#> 5  18.39684 18.39676 18.39668 18.39661 18.39653 18.39646 18.39640 18.39633
#> 6  15.81628 15.81622 15.81615 15.81609 15.81603 15.81597 15.81591 15.81586
#> 7  16.42483 16.42479 16.42474 16.42470 16.42466 16.42462 16.42458 16.42454
#> 8  12.91674 12.91686 12.91698 12.91708 12.91719 12.91729 12.91739 12.91748
#> 9  11.39389 11.39392 11.39395 11.39398 11.39400 11.39403 11.39405 11.39407
#> 10 10.92329 10.92323 10.92318 10.92314 10.92309 10.92304 10.92300 10.92296
#> 11 29.34457 29.34462 29.34467 29.34471 29.34476 29.34480 29.34485 29.34489
#> 12 15.26004 15.26002 15.26000 15.25998 15.25995 15.25993 15.25991 15.25989
#> 13 17.15476 17.15472 17.15469 17.15465 17.15462 17.15459 17.15456 17.15453
#> 14 13.93650 13.93639 13.93628 13.93618 13.93608 13.93599 13.93590 13.93581
#> 15 16.42179 16.42185 16.42191 16.42197 16.42203 16.42208 16.42214 16.42219
#> 16 26.17892 26.17895 26.17898 26.17900 26.17903 26.17905 26.17908 26.17910
#> 17 26.46499 26.46501 26.46504 26.46506 26.46509 26.46511 26.46513 26.46515
#> 18 16.78117 16.78110 16.78103 16.78097 16.78090 16.78084 16.78078 16.78073
#> 19 18.75346 18.75348 18.75349 18.75350 18.75352 18.75353 18.75354 18.75355
#> 20 14.29922 14.29929 14.29936 14.29942 14.29949 14.29955 14.29961 14.29967
#> 21 22.23699 22.23695 22.23691 22.23687 22.23684 22.23680 22.23677 22.23674
#>      0.0196   0.0183   0.0170   0.0159   0.0148   0.0138   0.0129   0.0120
#> 1  21.60929 21.60929 21.76592 21.76594 21.76595 21.76597 21.76599 21.76601
#> 2  23.54389 23.54389 23.57899 23.57904 23.57908 23.57912 23.57916 23.57920
#> 3  17.86633 17.86638 17.82826 17.82816 17.82807 17.82797 17.82788 17.82779
#> 4  15.67709 15.67714 15.65656 15.65636 15.65616 15.65598 15.65580 15.65563
#> 5  18.39627 18.39621 18.18160 18.18166 18.18171 18.18176 18.18181 18.18185
#> 6  15.81581 15.81576 15.84499 15.84511 15.84524 15.84535 15.84547 15.84558
#> 7  16.42451 16.42447 16.45737 16.45749 16.45759 16.45769 16.45779 16.45788
#> 8  12.91757 12.91766 12.92369 12.92352 12.92335 12.92318 12.92303 12.92288
#> 9  11.39410 11.39412 11.42024 11.42021 11.42017 11.42014 11.42011 11.42009
#> 10 10.92292 10.92288 10.96552 10.96560 10.96567 10.96575 10.96582 10.96589
#> 11 29.34493 29.34497 29.41030 29.41019 29.41009 29.40999 29.40990 29.40981
#> 12 15.25988 15.25986 15.23251 15.23256 15.23261 15.23265 15.23269 15.23274
#> 13 17.15450 17.15448 17.14860 17.14868 17.14875 17.14883 17.14890 17.14896
#> 14 13.93572 13.93564 13.95971 13.95977 13.95983 13.95989 13.95994 13.96000
#> 15 16.42224 16.42229 16.38218 16.38208 16.38199 16.38190 16.38181 16.38173
#> 16 26.17912 26.17914 26.21897 26.21897 26.21897 26.21896 26.21896 26.21896
#> 17 26.46517 26.46519 26.30692 26.30680 26.30669 26.30658 26.30648 26.30638
#> 18 16.78067 16.78062 16.77050 16.77074 16.77098 16.77120 16.77142 16.77162
#> 19 18.75356 18.75358 18.70957 18.70952 18.70946 18.70941 18.70937 18.70932
#> 20 14.29972 14.29977 14.32957 14.32950 14.32944 14.32938 14.32933 14.32927
#> 21 22.23671 22.23668 22.30803 22.30811 22.30819 22.30827 22.30834 22.30842
#>      0.0112   0.0105   0.0097   0.0091   0.0085   0.0079   0.0074   0.0069
#> 1  21.76603 21.76605 21.76607 21.76609 21.76611 21.76613 21.76615 21.76617
#> 2  23.57923 23.57927 23.57930 23.57933 23.57936 23.57939 23.57942 23.57944
#> 3  17.82770 17.82762 17.82753 17.82745 17.82738 17.82730 17.82723 17.82715
#> 4  15.65546 15.65531 15.65516 15.65501 15.65487 15.65474 15.65461 15.65449
#> 5  18.18190 18.18193 18.18197 18.18201 18.18205 18.18208 18.18211 18.18215
#> 6  15.84568 15.84578 15.84588 15.84597 15.84606 15.84615 15.84623 15.84631
#> 7  16.45797 16.45805 16.45813 16.45821 16.45829 16.45836 16.45843 16.45849
#> 8  12.92273 12.92259 12.92246 12.92233 12.92221 12.92209 12.92197 12.92186
#> 9  11.42006 11.42004 11.42002 11.41999 11.41997 11.41995 11.41993 11.41992
#> 10 10.96596 10.96603 10.96610 10.96616 10.96623 10.96629 10.96635 10.96641
#> 11 29.40972 29.40964 29.40956 29.40948 29.40941 29.40934 29.40927 29.40921
#> 12 15.23277 15.23281 15.23284 15.23288 15.23291 15.23294 15.23296 15.23299
#> 13 17.14902 17.14908 17.14914 17.14919 17.14924 17.14929 17.14934 17.14939
#> 14 13.96006 13.96012 13.96017 13.96023 13.96028 13.96034 13.96039 13.96044
#> 15 16.38164 16.38156 16.38149 16.38141 16.38134 16.38126 16.38119 16.38112
#> 16 26.21895 26.21895 26.21895 26.21894 26.21894 26.21893 26.21892 26.21892
#> 17 26.30629 26.30621 26.30612 26.30604 26.30597 26.30590 26.30583 26.30577
#> 18 16.77181 16.77199 16.77217 16.77234 16.77250 16.77265 16.77279 16.77293
#> 19 18.70928 18.70924 18.70920 18.70917 18.70914 18.70911 18.70908 18.70905
#> 20 14.32921 14.32916 14.32911 14.32906 14.32901 14.32896 14.32891 14.32887
#> 21 22.30849 22.30856 22.30863 22.30869 22.30876 22.30882 22.30888 22.30894
#>      0.0064   0.0060   0.0056   0.0052   0.0049   0.0045
#> 1  21.76619 21.76622 21.76658 21.81370 21.81392 21.81413
#> 2  23.57947 23.57951 23.57932 23.56500 23.56502 23.56504
#> 3  17.82708 17.82694 17.82692 17.81892 17.81898 17.81903
#> 4  15.65437 15.65414 15.65397 15.62358 15.62355 15.62354
#> 5  18.18218 18.18230 18.18210 18.21053 18.21048 18.21043
#> 6  15.84639 15.84654 15.84655 15.82449 15.82429 15.82410
#> 7  16.45856 16.45869 16.45862 16.42093 16.42065 16.42039
#> 8  12.92175 12.92155 12.92145 12.92817 12.92830 12.92842
#> 9  11.41990 11.41987 11.41986 11.42763 11.42767 11.42770
#> 10 10.96646 10.96657 10.96662 10.96673 10.96665 10.96656
#> 11 29.40914 29.40905 29.40889 29.41816 29.41824 29.41832
#> 12 15.23302 15.23304 15.23328 15.27112 15.27135 15.27156
#> 13 17.14943 17.14950 17.14966 17.16116 17.16118 17.16120
#> 14 13.96049 13.96059 13.96070 13.95753 13.95743 13.95733
#> 15 16.38106 16.38091 16.38096 16.39110 16.39124 16.39138
#> 16 26.21891 26.21890 26.21871 26.20753 26.20756 26.20758
#> 17 26.30571 26.30556 26.30583 26.27997 26.27963 26.27930
#> 18 16.77306 16.77329 16.77345 16.78668 16.78675 16.78682
#> 19 18.70902 18.70898 18.70918 18.73815 18.73825 18.73835
#> 20 14.32882 14.32876 14.32842 14.29002 14.29000 14.28997
#> 21 22.30900 22.30910 22.30894 22.29890 22.29886 22.29883
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 22.8 18.7 14.3 17.8 17.3 15.2 10.4 10.4 14.7 30.4 15.5 15.2 13.3 19.2
#> [16] 27.3 26.0 15.8 19.7 15.0 21.4
#> 
#> $X
#>               am       carb        cyl        disp        drat       gear
#>  [1,]  1.1547005  0.4673650 -0.4514661 -0.85531731  0.58642215  0.4291975
#>  [2,]  1.1547005 -1.3728847 -1.6365647 -1.26421605  0.49744191  0.4291975
#>  [3,] -0.8660254 -0.7594681  0.7336324  0.71737015 -0.74828144 -0.8583951
#>  [4,] -0.8660254  0.4673650  0.7336324  0.71737015 -0.64150515 -0.8583951
#>  [5,] -0.8660254  0.4673650 -0.4514661 -0.79555519  0.62201424  0.4291975
#>  [6,] -0.8660254 -0.1460516  0.7336324  0.05526873 -0.89064982 -0.8583951
#>  [7,] -0.8660254 -0.1460516  0.7336324  0.05526873 -0.89064982 -0.8583951
#>  [8,] -0.8660254  0.4673650  0.7336324  1.59807513 -1.13979449 -0.8583951
#>  [9,] -0.8660254  0.4673650  0.7336324  1.50371388 -1.01522216 -0.8583951
#> [10,] -0.8660254  0.4673650  0.7336324  1.34644514 -0.60591306 -0.8583951
#> [11,]  1.1547005 -0.7594681 -1.6365647 -1.51820507  2.41941507  0.4291975
#> [12,] -0.8660254 -0.7594681  0.7336324  0.38710579 -1.44232730 -0.8583951
#> [13,] -0.8660254 -0.7594681  0.7336324  0.27701766 -0.74828144 -0.8583951
#> [14,] -0.8660254  0.4673650  0.7336324  0.63873578  0.28388933 -0.8583951
#> [15,] -0.8660254 -0.7594681  0.7336324  1.03190764 -0.87285377 -0.8583951
#> [16,]  1.1547005 -1.3728847 -1.6365647 -1.49225573  0.90675101  0.4291975
#> [17,]  1.1547005 -0.7594681 -1.6365647 -1.16749577  1.52961268  1.7167902
#> [18,]  1.1547005  0.4673650  0.7336324  0.64659922  1.15589568  1.7167902
#> [19,]  1.1547005  1.6941981 -0.4514661 -0.97326887  0.08813281  1.7167902
#> [20,]  1.1547005  2.9210312  0.7336324  0.25342735 -0.05423557  1.7167902
#> [21,]  1.1547005 -0.7594681 -1.6365647 -1.16199136  0.96013915  0.4291975
#>               hp        qsec        vs          wt
#>  [1,] -0.8609188 -0.50590570 -0.559017 -0.77648402
#>  [2,] -1.1052336  1.13266262  1.788854 -1.07261874
#>  [3,]  0.0732260 -0.07911581 -0.559017  0.03295086
#>  [4,]  1.0792281 -0.97842308 -0.559017  0.16127591
#>  [5,] -0.6740898  1.35367881  1.788854  0.03295086
#>  [6,]  0.1450833  0.36291657 -0.559017  0.31921442
#>  [7,]  0.1450833  0.66776649 -0.559017  0.36857021
#>  [8,]  0.5043697  0.65252399 -0.559017  1.81963030
#>  [9,]  0.6480843  0.53058403 -0.559017  1.99138844
#> [10,]  0.8636562  0.22573411 -0.559017  1.91340630
#> [11,] -1.6944633  1.06407139  1.788854 -1.76853531
#> [12,] -0.2860604 -0.19343453 -0.559017  0.11192012
#> [13,] -0.2860604  0.13427913 -0.559017  0.02801528
#> [14,]  1.0792281 -1.30613674 -0.559017  0.42779715
#> [15,]  0.0732260 -0.05625207 -0.559017  0.43273273
#> [16,] -1.4932629  1.35367881  1.788854 -1.45265829
#> [17,] -1.1339765 -0.32299575 -0.559017 -1.25029956
#> [18,]  1.3522858 -1.99967031 -0.559017 -0.23357038
#> [19,]  0.0732260 -1.23754551 -0.559017 -0.62841667
#> [20,]  2.3726593 -1.92345783 -0.559017  0.16127591
#> [21,] -0.8752902  1.12504137  1.788854 -0.61854551
#> attr(,"center")
#>  [1]   0.4285714   3.2380952   6.7619048 268.7714286   3.5704762   3.6666667
#>  [7] 169.9047619  17.1238095   0.2380952   3.4066190
#> attr(,"scale")
#>  [1]   0.4948717   1.6302136   1.6876233 127.1708492   0.5619225   0.7766432
#>  [7]  69.5823628   1.3121211   0.4259177   1.0130525
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 40
#> 
#> $lambda.min
#> [1] 0.2977135
#> 
#> $null.dev
#> [1] 26.2195
#> 
#> $Bias
#> [1] 2.963478
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 16.67561