Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 41.345196 38.172565 33.343110 29.272253 25.949303 23.262175 21.114370
#>   [8] 19.422881 18.116373 17.133604 16.422057 15.936752 15.639220 15.496608
#>  [15] 15.480909 15.546450 15.197772 14.994789 14.994789 15.002425 15.015704
#>  [22] 14.908654 14.698098 14.228233 13.129511 12.862652 11.408865 10.683566
#>  [29] 10.190058  9.862914  9.703685  9.317573  8.025749  7.674955  8.056621
#>  [36]  8.314764  6.992226  5.931023  5.519965  4.271391  4.271254  4.271168
#>  [43]  4.271111  4.271070  4.261119  4.261070  4.174290  3.931019  4.127491
#>  [50]  4.284104  4.179948  4.349768  4.297326  3.575889  3.575777  3.705750
#>  [57]  3.705700  3.802499  3.802585  3.802634  3.855458  4.189896  4.222531
#>  [64]  4.160926  4.160881  4.147814  4.147742  4.147561  4.121518  4.121497
#>  [71]  4.136469  4.107146  4.091167  4.091424  4.622605  4.644514  4.796204
#>  [78]  4.813667  4.814203  4.814716  4.832818  4.833348  4.576109  4.576251
#>  [85]  4.576427  4.478657  5.076384  5.071400  5.105328  5.105493  5.105663
#>  [92]  5.105838  5.106026  5.105009  5.105170  5.105336  5.105414  5.109874
#>  [99]  5.109975  5.110013
#> 
#> $cvse
#>   [1] 11.5893293 11.2418041 10.0671323  9.0006695  8.0600095  7.2288320
#>   [7]  6.4957369  5.8535612  5.2987162  4.8304403  4.4498120  4.1583867
#>  [13]  3.9564865  3.8415226  3.8070184  3.8372350  3.7671534  3.7219796
#>  [19]  3.7219796  3.7183127  3.7148293  3.7126279  3.6737109  3.6052681
#>  [25]  3.4373050  3.3968486  3.3293765  3.2227028  3.1763088  3.1622551
#>  [31]  3.1498111  3.0248922  2.2705712  2.0582326  2.0010466  1.9521855
#>  [37]  1.3263633  1.1522245  1.1500755  1.0845527  1.0845604  1.0845586
#>  [43]  1.0845515  1.0845420  1.0738072  1.0738112  1.0938908  0.9567109
#>  [49]  0.9607539  0.9652484  0.9524341  0.9582951  0.9762440  0.9057039
#>  [55]  0.9056966  0.8923417  0.8923350  0.8901203  0.8901304  0.8901391
#>  [61]  0.9063384  0.9971268  1.0108096  1.0018315  1.0018855  1.0040048
#>  [67]  1.0040417  1.0040760  1.0305177  1.0305271  1.0334124  1.0446307
#>  [73]  1.0412145  1.0413009  1.1958084  1.2062316  1.2092235  1.2090730
#>  [79]  1.2092021  1.2093249  1.2228813  1.2230164  1.2309730  1.2311126
#>  [85]  1.2312496  1.2328466  1.3064609  1.3032261  1.2967765  1.2969426
#>  [91]  1.2971052  1.2972646  1.2974190  1.2978021  1.2979575  1.2981100
#>  [97]  1.2982142  1.3007846  1.3008802  1.3009302
#> 
#> $fold
#>  [1]  2  1  1  7  1  9  9  4  5 10  3  7  2  8  8  6  3  5  4  6 10
#> 
#> $lambda
#>   [1] 5.435820633 5.069464515 4.727799574 4.409161706 4.111998964 3.834863996
#>   [7] 3.576406998 3.335369136 3.110576419 2.900933979 2.705420738 2.523084435
#>  [13] 2.353036989 2.194450171 2.046551574 1.908620847 1.779986190 1.660021078
#>  [19] 1.548141214 1.443801677 1.346494277 1.255745069 1.171112054 1.092183021
#>  [25] 1.018573541 0.949925093 0.885903321 0.826196402 0.770513530 0.718583497
#>  [31] 0.670153374 0.624987280 0.582865229 0.543582063 0.506946452 0.472779957
#>  [37] 0.440916171 0.411199896 0.383486400 0.357640700 0.333536914 0.311057642
#>  [43] 0.290093398 0.270542074 0.252308444 0.235303699 0.219445017 0.204655157
#>  [49] 0.190862085 0.177998619 0.166002108 0.154814121 0.144380168 0.134649428
#>  [55] 0.125574507 0.117111205 0.109218302 0.101857354 0.094992509 0.088590332
#>  [61] 0.082619640 0.077051353 0.071858350 0.067015337 0.062498728 0.058286523
#>  [67] 0.054358206 0.050694645 0.047277996 0.044091617 0.041119990 0.038348640
#>  [73] 0.035764070 0.033353691 0.031105764 0.029009340 0.027054207 0.025230844
#>  [79] 0.023530370 0.021944502 0.020465516 0.019086208 0.017799862 0.016600211
#>  [85] 0.015481412 0.014438017 0.013464943 0.012557451 0.011711121 0.010921830
#>  [91] 0.010185735 0.009499251 0.008859033 0.008261964 0.007705135 0.007185835
#>  [97] 0.006701534 0.006249873 0.005828652 0.005435821
#> 
#> $fit
#> $beta
#>               5.4358     5.0695     4.7278     4.4092    4.1120    3.8349
#> (Intercept) 19.61429 21.4583994 23.3779143 25.1680603 26.837557 28.394534
#> am           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> cyl          0.00000 -0.2890029 -0.5898224 -0.8703677 -1.132005 -1.376009
#> disp         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> wt           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#>                3.5764   3.3354    3.1106    2.9009    2.7054    2.5231
#> (Intercept) 29.846577 31.20076 32.463669 33.641466 34.739883 35.764270
#> am           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.603568 -1.81579 -2.013709 -2.198289 -2.370429 -2.530968
#> disp         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#>                2.3530    2.1945    2.0466    1.9086    1.7800    1.6600
#> (Intercept) 36.719617 37.610577 38.441489 39.216401 39.830937 39.830937
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.680686 -2.820314 -2.950532 -3.071973 -3.168281 -3.168281
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.5481    1.4438    1.3465    1.2557     1.1711    1.0922
#> (Intercept) 39.830937 39.830937 39.830937 39.626817 39.0736403 38.557636
#> am           0.000000  0.000000  0.000000  0.164530  0.6104159  1.026339
#> carb         0.000000  0.000000  0.000000  0.000000  0.0000000  0.000000
#> cyl         -3.168281 -3.168281 -3.168281 -3.144887 -3.0814877 -3.022349
#> disp         0.000000  0.000000  0.000000  0.000000  0.0000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.0000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.0000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.0000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.0000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.0000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.0000000  0.000000
#>                1.0186    0.9499     0.8859    0.8262    0.7705    0.7186
#> (Intercept) 38.076409 36.539580 33.3143546 32.405637 32.081674 32.058655
#> am           1.414230  2.602396  5.0717784  5.756420  5.991325  6.008023
#> carb         0.000000 -0.275114 -0.9837052 -1.243798 -1.386414 -1.396508
#> cyl         -2.967195 -2.661126 -1.9568222 -1.729834 -1.625349 -1.617943
#> disp         0.000000  0.000000  0.0000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.0000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.0000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.0000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.0000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.0000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.0000000  0.000000  0.000000  0.000000
#>                0.6702    0.6250    0.5829    0.5436    0.5069    0.4728
#> (Intercept) 32.057987 32.057590 32.057354 32.057213 32.057130 32.057080
#> am           6.008536  6.008841  6.009022  6.009130  6.009194  6.009232
#> carb        -1.396648 -1.396731 -1.396781 -1.396810 -1.396828 -1.396838
#> cyl         -1.617801 -1.617716 -1.617665 -1.617635 -1.617618 -1.617607
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                 0.4409    0.4112    0.3835    0.3576    0.3335    0.3111
#> (Intercept) 17.5259096  1.997200  1.998311  1.998948  1.999302  1.999493
#> am           4.1185841  3.236578  3.236823  3.237003  3.237126  3.237208
#> carb        -2.1897066 -2.554138 -2.554195 -2.554221 -2.554230 -2.554232
#> cyl         -0.7867881  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.0000000  2.907522  2.907147  2.906964  2.906882  2.906851
#> gear         3.4162298  3.835377  3.835466  3.835475  3.835453  3.835425
#> hp           0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                0.2901    0.2705    0.2523    0.2353    0.2194    0.2047
#> (Intercept)  1.999590  1.999637  1.999658  1.999665  1.999666  1.999665
#> am           3.237261  3.237293  3.237313  3.237324  3.237331  3.237334
#> carb        -2.554230 -2.554228 -2.554226 -2.554225 -2.554224 -2.554223
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         2.906844  2.906847  2.906852  2.906858  2.906863  2.906866
#> gear         3.835399  3.835378  3.835363  3.835353  3.835347  3.835342
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                0.1909      0.1780      0.1660      0.1548      0.1444
#> (Intercept)  1.999664 -2.68695449 -2.68963000 -2.69183800 -2.69366016
#> am           3.237336  3.20129460  3.20098481  3.20072917  3.20051820
#> carb        -2.554222 -3.04845547 -3.04875338 -3.04899922 -3.04920210
#> cyl          0.000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         2.906869  3.70915994  3.70952625  3.70982858  3.71007809
#> gear         3.835340  4.11998829  4.12035800  4.12066308  4.12091483
#> hp           0.000000  0.01498724  0.01499402  0.01499961  0.01500423
#> qsec         0.000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.1346      0.1256      0.1171      0.1092      0.1019
#> (Intercept) -2.69516390 -2.69640487  0.40970350  0.41137313  0.41279929
#> am           3.20034410  3.20020042  4.11918114  4.12006678  4.12082981
#> carb        -3.04936952 -3.04950768 -2.82599859 -2.82588238 -2.82578341
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         3.71028401  3.71045395  3.16665244  3.16655545  3.16647480
#> gear         4.12112258  4.12129402  3.32634103  3.32561129  3.32498289
#> hp           0.01500804  0.01501119  0.01594039  0.01594261  0.01594456
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  1.32365101  1.32436561  1.32498039
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.0950      0.0886      0.0826      0.0771      0.0719
#> (Intercept)  0.41401655  0.41505470  0.41593940  0.41669272  0.41733363
#> am           4.12148677  4.12205200  4.12253798  4.12295552  4.12331402
#> carb        -2.82569921 -2.82562762 -2.82556680 -2.82551520 -2.82547144
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         3.16640787  3.16635245  3.16630666  3.16626893  3.16623792
#> gear         3.32444212  3.32397710  3.32357749  3.32323433  3.32293985
#> hp           0.01594627  0.01594776  0.01594906  0.01595019  0.01595118
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           1.32550894  1.32596303  1.32635288  1.32668734  1.32697408
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.0670     0.0625      0.0583      0.0544      0.0507
#> (Intercept)  3.14196893  3.1463754  3.15030525  3.15380464  3.15691619
#> am           3.90917032  3.9100218  3.91083319  3.91159950  3.91231794
#> carb        -2.81448758 -2.8141584 -2.81386443 -2.81360230 -2.81336892
#> cyl          0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> drat         3.14449658  3.1438872  3.14335219  3.14288303  3.14247207
#> gear         3.33058099  3.3297769  3.32902129  3.32831600  3.32766137
#> hp           0.01405748  0.0140539  0.01405101  0.01404869  0.01404684
#> qsec        -0.13716260 -0.1371887 -0.13721042 -0.13722837 -0.13724320
#> vs           1.55089555  1.5518153  1.55266848  1.55345598  1.55417970
#> wt           0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#>                  0.0473      0.0441       0.0411       0.0383       0.0358
#> (Intercept)  3.15967901  3.16212891  1.332156392  1.336144840  1.339878693
#> am           3.91298722  3.91360728  3.922837804  3.923902823  3.924880699
#> carb        -2.81316145 -2.81297727 -2.842502070 -2.842261993 -2.842040501
#> cyl          0.00000000  0.00000000  0.000000000  0.000000000  0.000000000
#> disp         0.00000000  0.00000000  0.004092531  0.004092834  0.004092774
#> drat         3.14211247  3.14179813  3.416865162  3.416327578  3.415827469
#> gear         3.32705680  3.32650094  3.600346031  3.599498679  3.598694986
#> hp           0.01404538  0.01404424  0.010745135  0.010744132  0.010743572
#> qsec        -0.13725541 -0.13726543 -0.175073439 -0.175099991 -0.175119839
#> vs           1.55484222  1.55544657  1.793655980  1.794689635  1.795613480
#> wt           0.00000000  0.00000000  0.000000000  0.000000000  0.000000000
#>                   0.0334       0.0311       0.0290       0.0271       0.0252
#> (Intercept)  1.343363878  1.346607743  0.143439623  0.146044648  0.148520934
#> am           3.925776372  3.926594755  3.999641441  4.000016269  4.000373274
#> carb        -2.841836680 -2.841649592 -2.528001342 -2.527695582 -2.527404821
#> cyl          0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> disp         0.004092411  0.004091798  0.010082580  0.010086922  0.010091058
#> drat         3.415363491  3.414934173  3.271860437  3.271563619  3.271281429
#> gear         3.597935486  3.597220204  3.360535766  3.360038632  3.359565825
#> hp           0.010743389  0.010743523  0.007324893  0.007320785  0.007316879
#> qsec        -0.175133892 -0.175142964  0.000000000  0.000000000  0.000000000
#> vs           1.796436677  1.797167847  1.955189182  1.955913002  1.956602089
#> wt           0.000000000  0.000000000 -0.771974496 -0.772448308 -0.772899277
#>                   0.0235       0.0219       0.0205       0.0191       0.0178
#> (Intercept) -2.053449837 -2.056544752 -2.059559551 -2.062495896 -2.065355442
#> am           4.224218886  4.224357152  4.224491910  4.224623248  4.224751250
#> carb        -2.641480323 -2.641727317 -2.641967681 -2.642201571 -2.642429145
#> cyl          0.260288954  0.260537011  0.260778605  0.261013873  0.261242956
#> disp         0.007623424  0.007618953  0.007614608  0.007610385  0.007606282
#> drat         3.388063698  3.388286619  3.388503751  3.388715217  3.388921134
#> gear         3.394912710  3.395105814  3.395293963  3.395477245  3.395655753
#> hp           0.008390083  0.008393461  0.008396745  0.008399937  0.008403040
#> qsec         0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> vs           2.242059836  2.242071052  2.242082683  2.242094681  2.242106998
#> wt          -0.594051453 -0.593655872 -0.593271229 -0.592897238 -0.592533622
#>                   0.0166       0.0155       0.0144       0.0135       0.0126
#> (Intercept) -2.068139837 -2.070850716 -2.073489704 -2.076058412 -2.078558437
#> am           4.224875996  4.224997565  4.225116030  4.225231463  4.225343935
#> carb        -2.642650556 -2.642865955 -2.643075489 -2.643279306 -2.643477547
#> cyl          0.261465989  0.261683108  0.261894448  0.262100141  0.262300317
#> disp         0.007602294  0.007598419  0.007594653  0.007590994  0.007587439
#> drat         3.389121623  3.389316802  3.389506789  3.389691702  3.389871655
#> gear         3.395829578  3.395998813  3.396163554  3.396323895  3.396479932
#> hp           0.008406056  0.008408987  0.008411836  0.008414605  0.008417297
#> qsec         0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> vs           2.242119588  2.242132408  2.242145416  2.242158571  2.242171838
#> wt          -0.592180106 -0.591836425 -0.591502316 -0.591177525 -0.590861802
#>                   0.0117       0.0109       0.0102       0.0095       0.0089
#> (Intercept) -3.478466776 -3.482497923 -3.486167454 -3.489500799 -3.492521921
#> am           4.328789441  4.329736567  4.330629657  4.331471219  4.332263692
#> carb        -2.631269868 -2.631109089 -2.630950744 -2.630795072 -2.630642283
#> cyl          0.333160952  0.333430652  0.333676025  0.333898777  0.334100514
#> disp         0.007644748  0.007647846  0.007650831  0.007653704  0.007656467
#> drat         3.400129145  3.400020636  3.399909058  3.399795095  3.399679371
#> gear         3.370955943  3.370435790  3.369933538  3.369449045  3.368982121
#> hp           0.008759307  0.008761122  0.008762812  0.008764382  0.008765839
#> qsec         0.060850675  0.061085506  0.061306634  0.061514759  0.061710550
#> vs           2.264467231  2.265202091  2.265892500  2.266540678  2.267148777
#> wt          -0.659616268 -0.660154490 -0.660669623 -0.661162404 -0.661633569
#>                   0.0083       0.0077       0.0072       0.0067      0.0062
#> (Intercept) -3.495253373 -3.497716367 -3.499930826 -3.501915451 -3.50368778
#> am           4.333009442  4.333710761  4.334369862  4.334988883  4.33556988
#> carb        -2.630492555 -2.630346043 -2.630202875 -2.630063158 -2.62992697
#> cyl          0.334282750  0.334446910  0.334594333  0.334726274  0.33484391
#> disp         0.007659124  0.007661678  0.007664129  0.007666483  0.00766874
#> drat         3.399562455  3.399444860  3.399327051  3.399209445  3.39909242
#> gear         3.368532536  3.368100022  3.367684278  3.367284976  3.36690176
#> hp           0.008767189  0.008768437  0.008769590  0.008770653  0.00877163
#> qsec         0.061894648  0.062067667  0.062230193  0.062382787  0.06252598
#> vs           2.267718878  2.268252987  2.268753035  2.269220875  2.26965829
#> wt          -0.662083846 -0.662513960 -0.662924624 -0.663316545 -0.66369042
#>                   0.0058       0.0054
#> (Intercept) -3.505264239 -3.506660204
#> am           4.336114840  4.336625660
#> carb        -2.629794390 -2.629665451
#> cyl          0.334948359  0.335040641
#> disp         0.007670906  0.007672981
#> drat         3.398976295  3.398861378
#> gear         3.366534268  3.366182101
#> hp           0.008772527  0.008773349
#> qsec         0.062660295  0.062786209
#> vs           2.270066973  2.270448563
#> wt          -0.664046924 -0.664386735
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   7   7   7   7  53  57  19  17   6   1   1   1   1   1   1
#>  [37]  58  26   1   1   1   1   1   1   1   1   1   1   1  45   1   1   1   1
#>  [55]   1  50   1   1   1   1   1   1   1  42   1   1   1   1   1   1  48   1
#>  [73]   1   1   1 145   1   1 100   1   1   1   1   1   1   1   1   1  52   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.435820633 5.069464515 4.727799574 4.409161706 4.111998964 3.834863996
#>   [7] 3.576406998 3.335369136 3.110576419 2.900933979 2.705420738 2.523084435
#>  [13] 2.353036989 2.194450171 2.046551574 1.908620847 1.779986190 1.660021078
#>  [19] 1.548141214 1.443801677 1.346494277 1.255745069 1.171112054 1.092183021
#>  [25] 1.018573541 0.949925093 0.885903321 0.826196402 0.770513530 0.718583497
#>  [31] 0.670153374 0.624987280 0.582865229 0.543582063 0.506946452 0.472779957
#>  [37] 0.440916171 0.411199896 0.383486400 0.357640700 0.333536914 0.311057642
#>  [43] 0.290093398 0.270542074 0.252308444 0.235303699 0.219445017 0.204655157
#>  [49] 0.190862085 0.177998619 0.166002108 0.154814121 0.144380168 0.134649428
#>  [55] 0.125574507 0.117111205 0.109218302 0.101857354 0.094992509 0.088590332
#>  [61] 0.082619640 0.077051353 0.071858350 0.067015337 0.062498728 0.058286523
#>  [67] 0.054358206 0.050694645 0.047277996 0.044091617 0.041119990 0.038348640
#>  [73] 0.035764070 0.033353691 0.031105764 0.029009340 0.027054207 0.025230844
#>  [79] 0.023530370 0.021944502 0.020465516 0.019086208 0.017799862 0.016600211
#>  [85] 0.015481412 0.014438017 0.013464943 0.012557451 0.011711121 0.010921830
#>  [91] 0.010185735 0.009499251 0.008859033 0.008261964 0.007705135 0.007185835
#>  [97] 0.006701534 0.006249873 0.005828652 0.005435821
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 25
#> 
#> $loss
#>   [1] 800.18571 693.65545 593.58421 510.19871 441.07954 384.13899 337.57673
#>   [8] 299.84139 269.59708 245.69447 227.14565 213.10228 202.83670 195.72552
#>  [15] 191.23539 188.91069 188.34534 188.34534 188.34534 188.34534 188.34534
#>  [22] 184.24557 174.06971 165.80870 159.17526 126.91479  83.42786  78.41349
#>  [29]  77.65560  77.65206  77.65206  77.65205  77.65205  77.65205  77.65205
#>  [36]  77.65205  56.70456  42.12866  42.12865  42.12865  42.12865  42.12865
#>  [43]  42.12865  42.12865  42.12865  42.12865  42.12865  42.12865  42.12865
#>  [50]  38.13334  38.13333  38.13332  38.13332  38.13331  38.13331  36.43792
#>  [57]  36.43790  36.43789  36.43789  36.43788  36.43788  36.43787  36.43787
#>  [64]  36.08045  36.08043  36.08041  36.08039  36.08038  36.08037  36.08036
#>  [71]  35.58515  35.58513  35.58511  35.58510  35.58508  35.14249  35.14246
#>  [78]  35.14244  34.91571  34.91569  34.91567  34.91565  34.91563  34.91562
#>  [85]  34.91560  34.91559  34.91557  34.91556  34.88593  34.88590  34.88587
#>  [92]  34.88585  34.88583  34.88581  34.88579  34.88577  34.88576  34.88575
#>  [99]  34.88574  34.88573
#> 
#> $linear.predictors
#>      5.4358   5.0695   4.7278   4.4092   4.1120   3.8349   3.5764   3.3354
#> 1  19.61429 19.72438 19.83898 19.94585 20.04553 20.13848 20.22517 20.30602
#> 2  19.61429 19.72438 19.83898 19.94585 20.04553 20.13848 20.22517 20.30602
#> 3  19.61429 19.72438 19.83898 19.94585 20.04553 20.13848 20.22517 20.30602
#> 4  19.61429 19.14638 18.65934 18.20512 17.78152 17.38646 17.01803 16.67443
#> 5  19.61429 19.14638 18.65934 18.20512 17.78152 17.38646 17.01803 16.67443
#> 6  19.61429 20.30239 21.01862 21.68659 22.30954 22.89050 23.43230 23.93760
#> 7  19.61429 19.72438 19.83898 19.94585 20.04553 20.13848 20.22517 20.30602
#> 8  19.61429 19.72438 19.83898 19.94585 20.04553 20.13848 20.22517 20.30602
#> 9  19.61429 19.14638 18.65934 18.20512 17.78152 17.38646 17.01803 16.67443
#> 10 19.61429 19.14638 18.65934 18.20512 17.78152 17.38646 17.01803 16.67443
#> 11 19.61429 19.14638 18.65934 18.20512 17.78152 17.38646 17.01803 16.67443
#> 12 19.61429 20.30239 21.01862 21.68659 22.30954 22.89050 23.43230 23.93760
#> 13 19.61429 20.30239 21.01862 21.68659 22.30954 22.89050 23.43230 23.93760
#> 14 19.61429 20.30239 21.01862 21.68659 22.30954 22.89050 23.43230 23.93760
#> 15 19.61429 19.14638 18.65934 18.20512 17.78152 17.38646 17.01803 16.67443
#> 16 19.61429 19.14638 18.65934 18.20512 17.78152 17.38646 17.01803 16.67443
#> 17 19.61429 19.14638 18.65934 18.20512 17.78152 17.38646 17.01803 16.67443
#> 18 19.61429 19.14638 18.65934 18.20512 17.78152 17.38646 17.01803 16.67443
#> 19 19.61429 20.30239 21.01862 21.68659 22.30954 22.89050 23.43230 23.93760
#> 20 19.61429 20.30239 21.01862 21.68659 22.30954 22.89050 23.43230 23.93760
#> 21 19.61429 19.14638 18.65934 18.20512 17.78152 17.38646 17.01803 16.67443
#>      3.1106   2.9009   2.7054   2.5231   2.3530   2.1945   2.0466   1.9086
#> 1  20.38141 20.45173 20.51731 20.57846 20.63550 20.68869 20.73830 20.78456
#> 2  20.38141 20.45173 20.51731 20.57846 20.63550 20.68869 20.73830 20.78456
#> 3  20.38141 20.45173 20.51731 20.57846 20.63550 20.68869 20.73830 20.78456
#> 4  16.35399 16.05515 15.77645 15.51653 15.27413 15.04806 14.83723 14.64061
#> 5  16.35399 16.05515 15.77645 15.51653 15.27413 15.04806 14.83723 14.64061
#> 6  24.40883 24.84831 25.25817 25.64040 25.99687 26.32932 26.63936 26.92851
#> 7  20.38141 20.45173 20.51731 20.57846 20.63550 20.68869 20.73830 20.78456
#> 8  20.38141 20.45173 20.51731 20.57846 20.63550 20.68869 20.73830 20.78456
#> 9  16.35399 16.05515 15.77645 15.51653 15.27413 15.04806 14.83723 14.64061
#> 10 16.35399 16.05515 15.77645 15.51653 15.27413 15.04806 14.83723 14.64061
#> 11 16.35399 16.05515 15.77645 15.51653 15.27413 15.04806 14.83723 14.64061
#> 12 24.40883 24.84831 25.25817 25.64040 25.99687 26.32932 26.63936 26.92851
#> 13 24.40883 24.84831 25.25817 25.64040 25.99687 26.32932 26.63936 26.92851
#> 14 24.40883 24.84831 25.25817 25.64040 25.99687 26.32932 26.63936 26.92851
#> 15 16.35399 16.05515 15.77645 15.51653 15.27413 15.04806 14.83723 14.64061
#> 16 16.35399 16.05515 15.77645 15.51653 15.27413 15.04806 14.83723 14.64061
#> 17 16.35399 16.05515 15.77645 15.51653 15.27413 15.04806 14.83723 14.64061
#> 18 16.35399 16.05515 15.77645 15.51653 15.27413 15.04806 14.83723 14.64061
#> 19 24.40883 24.84831 25.25817 25.64040 25.99687 26.32932 26.63936 26.92851
#> 20 24.40883 24.84831 25.25817 25.64040 25.99687 26.32932 26.63936 26.92851
#> 21 16.35399 16.05515 15.77645 15.51653 15.27413 15.04806 14.83723 14.64061
#>      1.7800   1.6600   1.5481   1.4438   1.3465   1.2557   1.1711   1.0922
#> 1  20.82125 20.82125 20.82125 20.82125 20.82125 20.92202 21.19513 21.44988
#> 2  20.82125 20.82125 20.82125 20.82125 20.82125 20.92202 21.19513 21.44988
#> 3  20.82125 20.82125 20.82125 20.82125 20.82125 20.75749 20.58471 20.42354
#> 4  14.48469 14.48469 14.48469 14.48469 14.48469 14.46772 14.42174 14.37885
#> 5  14.48469 14.48469 14.48469 14.48469 14.48469 14.46772 14.42174 14.37885
#> 6  27.15781 27.15781 27.15781 27.15781 27.15781 27.04727 26.74769 26.46824
#> 7  20.82125 20.82125 20.82125 20.82125 20.82125 20.75749 20.58471 20.42354
#> 8  20.82125 20.82125 20.82125 20.82125 20.82125 20.75749 20.58471 20.42354
#> 9  14.48469 14.48469 14.48469 14.48469 14.48469 14.46772 14.42174 14.37885
#> 10 14.48469 14.48469 14.48469 14.48469 14.48469 14.46772 14.42174 14.37885
#> 11 14.48469 14.48469 14.48469 14.48469 14.48469 14.46772 14.42174 14.37885
#> 12 27.15781 27.15781 27.15781 27.15781 27.15781 27.21180 27.35811 27.49458
#> 13 27.15781 27.15781 27.15781 27.15781 27.15781 27.21180 27.35811 27.49458
#> 14 27.15781 27.15781 27.15781 27.15781 27.15781 27.04727 26.74769 26.46824
#> 15 14.48469 14.48469 14.48469 14.48469 14.48469 14.46772 14.42174 14.37885
#> 16 14.48469 14.48469 14.48469 14.48469 14.48469 14.46772 14.42174 14.37885
#> 17 14.48469 14.48469 14.48469 14.48469 14.48469 14.46772 14.42174 14.37885
#> 18 14.48469 14.48469 14.48469 14.48469 14.48469 14.46772 14.42174 14.37885
#> 19 27.15781 27.15781 27.15781 27.15781 27.15781 27.21180 27.35811 27.49458
#> 20 27.15781 27.15781 27.15781 27.15781 27.15781 27.21180 27.35811 27.49458
#> 21 14.48469 14.48469 14.48469 14.48469 14.48469 14.63225 15.03215 15.40519
#>      1.0186   0.9499   0.8859   0.8262   0.7705   0.7186   0.6702   0.6250
#> 1  21.68747 22.07476 22.71038 22.80786 22.77525 22.77299 22.77313 22.77321
#> 2  21.68747 22.07476 22.71038 22.80786 22.77525 22.77299 22.77313 22.77321
#> 3  20.27324 20.29771 20.58972 20.78283 20.94317 20.95449 20.95453 20.95456
#> 4  14.33885 14.70034 15.69237 16.07936 16.30605 16.32209 16.32228 16.32240
#> 5  14.33885 14.15012 13.72496 13.59177 13.53323 13.52908 13.52899 13.52894
#> 6  26.20763 25.34485 23.51966 22.99870 22.80745 22.79387 22.79349 22.79326
#> 7  20.27324 19.47237 17.63860 17.05144 16.78392 16.76496 16.76459 16.76437
#> 8  20.27324 19.47237 17.63860 17.05144 16.78392 16.76496 16.76459 16.76437
#> 9  14.33885 14.15012 13.72496 13.59177 13.53323 13.52908 13.52899 13.52894
#> 10 14.33885 14.15012 13.72496 13.59177 13.53323 13.52908 13.52899 13.52894
#> 11 14.33885 14.15012 13.72496 13.59177 13.53323 13.52908 13.52899 13.52894
#> 12 27.62186 28.22236 29.57514 29.99892 30.18519 30.19840 30.19867 30.19884
#> 13 27.62186 27.94724 28.59143 28.75512 28.79878 28.80189 28.80202 28.80210
#> 14 26.20763 25.61996 24.50336 24.24250 24.19386 24.19037 24.19014 24.18999
#> 15 14.33885 14.70034 15.69237 16.07936 16.30605 16.32209 16.32228 16.32240
#> 16 14.33885 14.70034 15.69237 16.07936 16.30605 16.32209 16.32228 16.32240
#> 17 14.33885 14.15012 13.72496 13.59177 13.53323 13.52908 13.52899 13.52894
#> 18 14.33885 14.70034 15.69237 16.07936 16.30605 16.32209 16.32228 16.32240
#> 19 27.62186 28.22236 29.57514 29.99892 30.18519 30.19840 30.19867 30.19884
#> 20 27.62186 27.94724 28.59143 28.75512 28.79878 28.80189 28.80202 28.80210
#> 21 15.75307 15.65206 14.86191 14.37299 13.97890 13.95107 13.95093 13.95085
#>      0.5829   0.5436   0.5069   0.4728   0.4409   0.4112   0.3835   0.3576
#> 1  22.77326 22.77329 22.77331 22.77332 21.82986 21.69807 21.69809 21.69813
#> 2  22.77326 22.77329 22.77331 22.77332 21.82986 21.69807 21.69809 21.69813
#> 3  20.95458 20.95459 20.95460 20.95460 20.86416 19.90436 19.90453 19.90460
#> 4  16.32247 16.32251 16.32253 16.32255 17.10088 17.55375 17.55383 17.55387
#> 5  13.52891 13.52889 13.52888 13.52887 12.72147 12.61993 12.61987 12.61984
#> 6  22.79313 22.79305 22.79300 22.79298 23.66426 23.62792 23.62780 23.62770
#> 7  16.76424 16.76416 16.76411 16.76409 17.71127 18.51964 18.51941 18.51926
#> 8  16.76424 16.76416 16.76411 16.76409 17.71127 18.51964 18.51941 18.51926
#> 9  13.52891 13.52889 13.52888 13.52887 12.72147 11.80582 11.80587 11.80589
#> 10 13.52891 13.52889 13.52888 13.52887 12.72147 12.00935 12.00937 12.00938
#> 11 13.52891 13.52889 13.52888 13.52887 12.72147 12.67808 12.67801 12.67798
#> 12 30.19893 30.19899 30.19903 30.19905 29.97255 29.88384 29.88396 29.88404
#> 13 28.80215 28.80218 28.80220 28.80221 27.78285 29.80110 29.80084 29.80074
#> 14 24.18991 24.18986 24.18983 24.18981 22.43774 21.70703 21.70696 21.70692
#> 15 16.32247 16.32251 16.32253 16.32255 17.10088 16.41982 16.42004 16.42015
#> 16 16.32247 16.32251 16.32253 16.32255 17.10088 17.55375 17.55383 17.55387
#> 17 13.52891 13.52889 13.52888 13.52887 12.72147 14.13184 14.13159 14.13146
#> 18 16.32247 16.32251 16.32253 16.32255 17.10088 17.35022 17.35033 17.35038
#> 19 30.19893 30.19899 30.19903 30.19905 29.97255 29.88384 29.88396 29.88404
#> 20 28.80215 28.80218 28.80220 28.80221 31.19908 30.26375 30.26402 30.26414
#> 21 13.95081 13.95078 13.95076 13.95075 14.91368 14.27019 14.27020 14.27021
#>      0.3335   0.3111   0.2901   0.2705   0.2523   0.2353   0.2194   0.2047
#> 1  21.69816 21.69819 21.69822 21.69823 21.69824 21.69825 21.69825 21.69826
#> 2  21.69816 21.69819 21.69822 21.69823 21.69824 21.69825 21.69825 21.69826
#> 3  19.90463 19.90464 19.90464 19.90463 19.90463 19.90462 19.90462 19.90462
#> 4  17.55388 17.55388 17.55388 17.55388 17.55388 17.55388 17.55388 17.55388
#> 5  12.61983 12.61983 12.61983 12.61984 12.61984 12.61984 12.61984 12.61984
#> 6  23.62763 23.62758 23.62755 23.62753 23.62752 23.62751 23.62751 23.62750
#> 7  18.51917 18.51912 18.51909 18.51908 18.51907 18.51906 18.51906 18.51906
#> 8  18.51917 18.51912 18.51909 18.51908 18.51907 18.51906 18.51906 18.51906
#> 9  11.80591 11.80591 11.80592 11.80592 11.80592 11.80592 11.80592 11.80592
#> 10 12.00939 12.00939 12.00940 12.00940 12.00940 12.00940 12.00940 12.00940
#> 11 12.67797 12.67797 12.67797 12.67797 12.67798 12.67798 12.67798 12.67798
#> 12 29.88409 29.88412 29.88414 29.88415 29.88415 29.88416 29.88416 29.88416
#> 13 29.80071 29.80071 29.80073 29.80074 29.80075 29.80076 29.80077 29.80077
#> 14 21.70690 21.70688 21.70688 21.70688 21.70687 21.70687 21.70687 21.70687
#> 15 16.42020 16.42021 16.42021 16.42021 16.42021 16.42020 16.42020 16.42020
#> 16 17.55388 17.55388 17.55388 17.55388 17.55388 17.55388 17.55388 17.55388
#> 17 14.13141 14.13140 14.13139 14.13140 14.13140 14.13141 14.13141 14.13141
#> 18 17.35040 17.35041 17.35041 17.35040 17.35040 17.35040 17.35040 17.35039
#> 19 29.88409 29.88412 29.88414 29.88415 29.88415 29.88416 29.88416 29.88416
#> 20 30.26418 30.26419 30.26419 30.26418 30.26417 30.26416 30.26415 30.26415
#> 21 14.27022 14.27023 14.27023 14.27023 14.27023 14.27023 14.27024 14.27024
#>      0.1909   0.1780   0.1660   0.1548   0.1444   0.1346   0.1256   0.1171
#> 1  21.69826 20.91479 20.91427 20.91384 20.91348 20.91318 20.91294 20.63364
#> 2  21.69826 20.91479 20.91427 20.91384 20.91348 20.91318 20.91294 20.63364
#> 3  19.90462 19.69736 19.69737 19.69738 19.69739 19.69739 19.69740 20.39311
#> 4  17.55387 17.88272 17.88290 17.88305 17.88317 17.88327 17.88335 17.50125
#> 5  12.61984 13.05747 13.05754 13.05761 13.05766 13.05771 13.05774 13.15508
#> 6  23.62750 23.65978 23.66007 23.66031 23.66050 23.66066 23.66080 23.31434
#> 7  18.51906 17.98251 17.98240 17.98230 17.98222 17.98215 17.98210 18.10867
#> 8  18.51906 17.98251 17.98240 17.98230 17.98222 17.98215 17.98210 18.10867
#> 9  11.80592 11.41941 11.41912 11.41887 11.41867 11.41851 11.41837 11.63080
#> 10 12.00940 11.82892 11.82872 11.82856 11.82842 11.82831 11.82821 12.01187
#> 11 12.67798 12.90684 12.90682 12.90681 12.90680 12.90679 12.90679 12.97931
#> 12 29.88416 30.06837 30.06851 30.06862 30.06871 30.06879 30.06885 30.30391
#> 13 29.80078 29.96288 29.96293 29.96298 29.96302 29.96305 29.96308 29.94640
#> 14 21.70687 21.80221 21.80236 21.80248 21.80258 21.80267 21.80273 22.14921
#> 15 16.42020 16.06147 16.06133 16.06122 16.06113 16.06106 16.06099 15.86775
#> 16 17.55387 17.50804 17.50805 17.50805 17.50806 17.50807 17.50807 17.10274
#> 17 14.13141 14.98623 14.98650 14.98672 14.98690 14.98706 14.98718 14.80174
#> 18 17.35039 17.62308 17.62323 17.62336 17.62346 17.62355 17.62362 17.27959
#> 19 29.88416 30.06837 30.06851 30.06862 30.06871 30.06879 30.06885 30.30391
#> 20 30.26415 30.69446 30.69488 30.69522 30.69550 30.69573 30.69593 30.57179
#> 21 14.27024 14.87779 14.87784 14.87788 14.87791 14.87794 14.87796 15.10258
#>      0.1092   0.1019   0.0950   0.0886   0.0826   0.0771   0.0719   0.0670
#> 1  20.63361 20.63358 20.63356 20.63353 20.63351 20.63349 20.63348 20.66768
#> 2  20.63361 20.63358 20.63356 20.63353 20.63351 20.63349 20.63348 20.59086
#> 3  20.39337 20.39359 20.39378 20.39394 20.39408 20.39420 20.39430 20.43505
#> 4  17.50105 17.50088 17.50073 17.50060 17.50049 17.50040 17.50032 17.53545
#> 5  13.15526 13.15542 13.15555 13.15567 13.15577 13.15585 13.15593 13.24102
#> 6  23.31386 23.31346 23.31311 23.31281 23.31255 23.31233 23.31214 22.90708
#> 7  18.10849 18.10834 18.10821 18.10809 18.10799 18.10790 18.10783 18.30266
#> 8  18.10849 18.10834 18.10821 18.10809 18.10799 18.10790 18.10783 18.22036
#> 9  11.63092 11.63102 11.63111 11.63118 11.63124 11.63129 11.63134 11.50474
#> 10 12.01201 12.01212 12.01222 12.01230 12.01237 12.01243 12.01248 11.88737
#> 11 12.97945 12.97958 12.97969 12.97978 12.97986 12.97993 12.97999 12.87633
#> 12 30.30413 30.30432 30.30448 30.30462 30.30474 30.30485 30.30494 30.19665
#> 13 29.94662 29.94681 29.94698 29.94712 29.94725 29.94736 29.94745 29.98849
#> 14 22.14938 22.14952 22.14965 22.14976 22.14985 22.14993 22.15000 22.12371
#> 15 15.86753 15.86734 15.86717 15.86703 15.86690 15.86680 15.86670 15.97824
#> 16 17.10248 17.10226 17.10207 17.10190 17.10176 17.10164 17.10154 17.14561
#> 17 14.80187 14.80198 14.80208 14.80217 14.80225 14.80231 14.80237 14.93514
#> 18 17.27939 17.27922 17.27908 17.27895 17.27885 17.27876 17.27868 17.31122
#> 19 30.30413 30.30432 30.30448 30.30462 30.30474 30.30485 30.30494 30.27484
#> 20 30.57153 30.57130 30.57111 30.57095 30.57080 30.57068 30.57058 30.75116
#> 21 15.10282 15.10303 15.10320 15.10336 15.10349 15.10360 15.10370 15.02634
#>      0.0625   0.0583   0.0544   0.0507   0.0473   0.0441   0.0411   0.0383
#> 1  20.66783 20.66797 20.66808 20.66817 20.66825 20.66832 20.56721 20.56723
#> 2  20.59101 20.59113 20.59123 20.59132 20.59139 20.59145 20.46916 20.46918
#> 3  20.43552 20.43594 20.43632 20.43667 20.43698 20.43727 20.44270 20.44322
#> 4  17.53511 17.53481 17.53452 17.53427 17.53404 17.53383 17.58528 17.58499
#> 5  13.24109 13.24116 13.24123 13.24130 13.24138 13.24145 13.06403 13.06415
#> 6  22.90652 22.90600 22.90552 22.90507 22.90467 22.90430 22.82414 22.82348
#> 7  18.30278 18.30287 18.30293 18.30297 18.30299 18.30300 18.35502 18.35494
#> 8  18.22046 18.22054 18.22059 18.22062 18.22064 18.22064 18.24997 18.24988
#> 9  11.50506 11.50535 11.50561 11.50584 11.50604 11.50623 11.76121 11.76150
#> 10 11.88762 11.88785 11.88805 11.88824 11.88841 11.88856 12.08674 12.08699
#> 11 12.87640 12.87647 12.87654 12.87661 12.87668 12.87674 13.02198 13.02209
#> 12 30.19671 30.19678 30.19686 30.19694 30.19703 30.19711 30.17092 30.17111
#> 13 29.98843 29.98841 29.98840 29.98840 29.98842 29.98844 30.23637 30.23637
#> 14 22.12383 22.12394 22.12406 22.12417 22.12427 22.12436 21.75732 21.75746
#> 15 15.97823 15.97820 15.97817 15.97812 15.97807 15.97801 15.83844 15.83839
#> 16 17.14535 17.14511 17.14488 17.14467 17.14447 17.14429 17.03844 17.03816
#> 17 14.93490 14.93470 14.93454 14.93440 14.93430 14.93421 14.87516 14.87501
#> 18 17.31093 17.31065 17.31041 17.31018 17.30997 17.30978 17.50454 17.50431
#> 19 30.27491 30.27499 30.27508 30.27517 30.27526 30.27535 30.27194 30.27214
#> 20 30.75084 30.75053 30.75025 30.74999 30.74975 30.74953 30.89162 30.89139
#> 21 15.02648 15.02661 15.02675 15.02688 15.02701 15.02713 14.88781 14.88801
#>      0.0358   0.0334   0.0311   0.0290   0.0271   0.0252   0.0235   0.0219
#> 1  20.56727 20.56731 20.56735 20.62985 20.62991 20.62997 20.54592 20.54580
#> 2  20.46920 20.46923 20.46927 20.43300 20.43294 20.43288 20.39444 20.39442
#> 3  20.44368 20.44410 20.44448 20.65471 20.65508 20.65544 20.70871 20.70863
#> 4  17.58472 17.58447 17.58424 17.73140 17.73140 17.73141 17.77220 17.77222
#> 5  13.06428 13.06442 13.06455 13.28409 13.28434 13.28458 13.20260 13.20243
#> 6  22.82289 22.82236 22.82189 22.99423 22.99375 22.99329 22.80684 22.80684
#> 7  18.35487 18.35480 18.35472 18.18967 18.18966 18.18965 18.31141 18.31150
#> 8  18.24980 18.24972 18.24964 18.18967 18.18966 18.18965 18.31141 18.31150
#> 9  11.76174 11.76194 11.76211 11.90731 11.90749 11.90767 11.77416 11.77395
#> 10 12.08719 12.08737 12.08751 11.95427 11.95426 11.95425 11.90038 11.90034
#> 11 13.02218 13.02226 13.02232 12.67601 12.67582 12.67564 12.69994 12.70007
#> 12 30.17128 30.17144 30.17160 29.94020 29.94004 29.93988 30.06225 30.06247
#> 13 30.23637 30.23636 30.23634 30.51209 30.51230 30.51251 30.50781 30.50771
#> 14 21.75762 21.75779 21.75798 21.77663 21.77663 21.77663 21.57393 21.57377
#> 15 15.83833 15.83829 15.83825 15.78702 15.78703 15.78703 15.87340 15.87346
#> 16 17.03790 17.03768 17.03747 16.98751 16.98738 16.98725 17.13851 17.13869
#> 17 14.87489 14.87479 14.87471 14.67620 14.67612 14.67605 14.72777 14.72786
#> 18 17.50407 17.50384 17.50361 17.59302 17.59303 17.59303 17.59938 17.59937
#> 19 30.27233 30.27250 30.27266 30.14780 30.14776 30.14773 30.22196 30.22208
#> 20 30.89116 30.89093 30.89071 30.79843 30.79837 30.79832 30.69285 30.69277
#> 21 14.88821 14.88841 14.88860 15.03688 15.03701 15.03713 15.07411 15.07412
#>      0.0205   0.0191   0.0178   0.0166   0.0155   0.0144   0.0135   0.0126
#> 1  20.54568 20.54557 20.54545 20.54534 20.54524 20.54513 20.54503 20.54493
#> 2  20.39440 20.39438 20.39436 20.39434 20.39432 20.39430 20.39428 20.39426
#> 3  20.70854 20.70846 20.70838 20.70831 20.70823 20.70816 20.70809 20.70802
#> 4  17.77224 17.77226 17.77228 17.77229 17.77231 17.77233 17.77235 17.77236
#> 5  13.20226 13.20210 13.20194 13.20178 13.20163 13.20148 13.20134 13.20120
#> 6  22.80684 22.80683 22.80683 22.80683 22.80683 22.80682 22.80682 22.80682
#> 7  18.31159 18.31167 18.31176 18.31184 18.31192 18.31199 18.31207 18.31214
#> 8  18.31159 18.31167 18.31176 18.31184 18.31192 18.31199 18.31207 18.31214
#> 9  11.77375 11.77355 11.77336 11.77318 11.77300 11.77283 11.77266 11.77249
#> 10 11.90031 11.90028 11.90024 11.90021 11.90018 11.90015 11.90012 11.90010
#> 11 12.70019 12.70031 12.70043 12.70054 12.70065 12.70075 12.70085 12.70095
#> 12 30.06269 30.06290 30.06311 30.06331 30.06350 30.06369 30.06388 30.06405
#> 13 30.50762 30.50752 30.50743 30.50734 30.50726 30.50718 30.50710 30.50702
#> 14 21.57360 21.57344 21.57328 21.57313 21.57298 21.57284 21.57269 21.57256
#> 15 15.87353 15.87359 15.87365 15.87372 15.87377 15.87383 15.87389 15.87394
#> 16 17.13887 17.13904 17.13921 17.13938 17.13954 17.13969 17.13984 17.13999
#> 17 14.72795 14.72804 14.72813 14.72821 14.72829 14.72837 14.72845 14.72853
#> 18 17.59935 17.59934 17.59933 17.59932 17.59930 17.59929 17.59928 17.59927
#> 19 30.22219 30.22230 30.22241 30.22252 30.22262 30.22272 30.22281 30.22291
#> 20 30.69269 30.69260 30.69253 30.69245 30.69237 30.69230 30.69223 30.69217
#> 21 15.07412 15.07413 15.07413 15.07413 15.07414 15.07414 15.07415 15.07415
#>      0.0117   0.0109   0.0102   0.0095   0.0089   0.0083   0.0077   0.0072
#> 1  20.52863 20.52845 20.52829 20.52814 20.52801 20.52788 20.52777 20.52767
#> 2  20.39450 20.39432 20.39415 20.39399 20.39385 20.39371 20.39359 20.39347
#> 3  20.73710 20.73752 20.73792 20.73830 20.73866 20.73900 20.73932 20.73962
#> 4  17.79914 17.79927 17.79938 17.79948 17.79958 17.79966 17.79974 17.79981
#> 5  13.19621 13.19643 13.19664 13.19685 13.19705 13.19725 13.19744 13.19763
#> 6  22.89264 22.89253 22.89242 22.89232 22.89223 22.89214 22.89205 22.89197
#> 7  18.27536 18.27501 18.27467 18.27435 18.27404 18.27374 18.27346 18.27320
#> 8  18.31187 18.31166 18.31145 18.31126 18.31106 18.31088 18.31070 18.31053
#> 9  11.77208 11.77220 11.77232 11.77243 11.77255 11.77266 11.77277 11.77287
#> 10 11.88143 11.88140 11.88136 11.88134 11.88131 11.88129 11.88127 11.88125
#> 11 12.66973 12.66958 12.66944 12.66931 12.66918 12.66906 12.66894 12.66883
#> 12 30.08588 30.08600 30.08610 30.08620 30.08629 30.08637 30.08644 30.08651
#> 13 30.52722 30.52746 30.52769 30.52791 30.52812 30.52831 30.52849 30.52866
#> 14 21.54017 21.54008 21.54000 21.53993 21.53988 21.53984 21.53981 21.53979
#> 15 15.87113 15.87105 15.87096 15.87088 15.87080 15.87072 15.87065 15.87057
#> 16 17.17239 17.17236 17.17233 17.17230 17.17226 17.17221 17.17217 17.17212
#> 17 14.68357 14.68345 14.68334 14.68324 14.68314 14.68305 14.68297 14.68289
#> 18 17.60160 17.60165 17.60169 17.60172 17.60175 17.60177 17.60179 17.60181
#> 19 30.22828 30.22841 30.22853 30.22864 30.22875 30.22884 30.22893 30.22901
#> 20 30.60535 30.60505 30.60476 30.60449 30.60424 30.60400 30.60377 30.60356
#> 21 15.12571 15.12614 15.12654 15.12692 15.12728 15.12762 15.12794 15.12825
#>      0.0067   0.0062   0.0058   0.0054
#> 1  20.52758 20.52749 20.52742 20.52735
#> 2  20.39336 20.39327 20.39317 20.39309
#> 3  20.73990 20.74017 20.74043 20.74066
#> 4  17.79988 17.79994 17.79999 17.80003
#> 5  13.19781 13.19798 13.19815 13.19831
#> 6  22.89190 22.89183 22.89176 22.89169
#> 7  18.27294 18.27270 18.27247 18.27225
#> 8  18.31037 18.31021 18.31006 18.30992
#> 9  11.77298 11.77307 11.77317 11.77326
#> 10 11.88123 11.88122 11.88120 11.88119
#> 11 12.66873 12.66863 12.66853 12.66844
#> 12 30.08656 30.08661 30.08666 30.08670
#> 13 30.52882 30.52896 30.52910 30.52924
#> 14 21.53978 21.53978 21.53978 21.53980
#> 15 15.87050 15.87044 15.87037 15.87031
#> 16 17.17207 17.17202 17.17197 17.17192
#> 17 14.68281 14.68274 14.68267 14.68261
#> 18 17.60182 17.60183 17.60184 17.60184
#> 19 30.22908 30.22915 30.22921 30.22927
#> 20 30.60336 30.60317 30.60299 30.60282
#> 21 15.12853 15.12881 15.12906 15.12930
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 21.4 18.7 14.3 22.8 19.2 17.8 10.4 10.4 14.7 32.4 30.4 21.5 15.5
#> [16] 15.2 13.3 19.2 27.3 30.4 15.0
#> 
#> $X
#>               am       carb        cyl        disp        drat       gear
#>  [1,]  1.4142136  0.6361464 -0.2236068 -0.68628723  0.64346846  0.6495191
#>  [2,]  1.4142136  0.6361464 -0.2236068 -0.68628723  0.64346846  0.6495191
#>  [3,] -0.7071068 -1.1855456 -0.2236068  0.05410862 -0.95776375 -0.8660254
#>  [4,] -0.7071068 -0.5783149  0.9503289  0.82472471 -0.82107320 -0.8660254
#>  [5,] -0.7071068  0.6361464  0.9503289  0.82472471 -0.70390987 -0.8660254
#>  [6,] -0.7071068 -0.5783149 -1.3975425 -0.83134438  0.68252291  0.6495191
#>  [7,] -0.7071068  0.6361464 -0.2236068 -0.62886878  0.68252291  0.6495191
#>  [8,] -0.7071068  0.6361464 -0.2236068 -0.62886878  0.68252291  0.6495191
#>  [9,] -0.7071068  0.6361464  0.9503289  1.67089140 -1.25067209 -0.8660254
#> [10,] -0.7071068  0.6361464  0.9503289  1.58023068 -1.11398153 -0.8660254
#> [11,] -0.7071068  0.6361464  0.9503289  1.42912949 -0.66485542 -0.8660254
#> [12,]  1.4142136 -1.1855456 -1.3975425 -1.30051359  0.99495846  0.6495191
#> [13,]  1.4142136 -0.5783149 -1.3975425 -1.32317877  2.65477235  0.6495191
#> [14,] -0.7071068 -1.1855456 -1.3975425 -0.98773412  0.25292402 -0.8660254
#> [15,] -0.7071068 -0.5783149  0.9503289  0.50741220 -1.58263486 -0.8660254
#> [16,] -0.7071068 -0.5783149  0.9503289  0.40164136 -0.82107320 -0.8660254
#> [17,] -0.7071068  0.6361464  0.9503289  0.74917411  0.31150569 -0.8660254
#> [18,] -0.7071068 -0.5783149  0.9503289  1.12692710 -0.95776375 -0.8660254
#> [19,]  1.4142136 -1.1855456 -1.3975425 -1.29824707  0.99495846  0.6495191
#> [20,]  1.4142136 -0.5783149 -1.3975425 -1.17661061  0.38961458  2.1650635
#> [21,]  1.4142136  3.0650691  0.9503289  0.37897619 -0.05951153  2.1650635
#>                hp         qsec         vs          wt
#>  [1,] -0.59467713 -0.770428930 -0.8660254 -0.65662775
#>  [2,] -0.59467713 -0.452190637 -0.8660254 -0.41721592
#>  [3,] -0.59467713  0.923053418  1.1547005 -0.09800014
#>  [4,]  0.32774819 -0.452190637 -0.8660254  0.11324560
#>  [5,]  1.32112931 -1.122764184 -0.8660254  0.23529869
#>  [6,] -0.80754452  2.889311446  1.1547005 -0.15902668
#>  [7,] -0.41019207  0.275211177  1.1547005  0.11324560
#>  [8,] -0.41019207  0.616180778  1.1547005  0.11324560
#>  [9,]  0.75348296  0.093360724 -0.8660254  1.81260019
#> [10,]  0.89539455  0.002435497 -0.8660254  1.97596356
#> [11,]  1.10826193 -0.224877570 -0.8660254  1.90179283
#> [12,] -1.21908812  0.940101898  1.1547005 -1.05095313
#> [13,] -1.41776435  0.400233364  1.1547005 -1.60019204
#> [14,] -0.77916220  1.246974538  1.1547005 -0.80215259
#> [15,] -0.02703078 -0.537433037 -0.8660254  0.18835519
#> [16,] -0.02703078 -0.293071490 -0.8660254  0.10855125
#> [17,]  1.32112931 -1.367125731 -0.8660254  0.48879357
#> [18,]  0.32774819 -0.435142157 -0.8660254  0.49348792
#> [19,] -1.21908812  0.616180778  1.1547005 -1.29975366
#> [20,] -0.55210366 -0.520384557  1.1547005 -1.69595677
#> [21,]  2.59833361 -1.827434691 -0.8660254  0.23529869
#> attr(,"center")
#>  [1]   0.3333333   2.9523810   6.3809524 250.8380952   3.5704762   3.5714286
#>  [7] 151.9047619  17.8157143   0.4285714   3.3193810
#> attr(,"scale")
#>  [1]   0.4714045   1.6468206   1.7036708 132.3616276   0.5121056   0.6598289
#>  [7]  70.4664086   1.7596877   0.4948717   1.0651103
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 55
#> 
#> $lambda.min
#> [1] 0.1255745
#> 
#> $null.dev
#> [1] 38.10408
#> 
#> $Bias
#> [1] 1.207837
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 29.34046