Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 40.26820 35.84096 30.77680 26.70223 23.46610 20.93848 19.00772 17.57806
#>   [9] 16.56741 15.90549 15.53225 15.39645 15.45443 15.66908 16.00894 16.42823
#>  [17] 16.33765 16.06610 15.77203 15.51376 15.28681 15.08727 14.91170 14.81127
#>  [25] 16.57982 16.86909 16.87442 16.21088 16.13089 16.06429 16.64508 16.44080
#>  [33] 17.06392 17.10050 17.06875 17.04264 17.01429 19.86537 19.82629 19.79049
#>  [41] 19.75832 21.44051 21.40482 21.74023 23.36387 34.06778 36.74581 36.74875
#>  [49] 37.78101 37.19029 39.91009 39.88383 42.90758 48.65163 57.95597 58.30898
#>  [57] 58.90604 59.01246 59.23345 57.08438 56.94546 56.62289 59.97499 63.41915
#>  [65] 63.42179 62.16693 61.70005 61.72262 61.48174 61.48306 61.48428 61.48540
#>  [73] 59.66581 59.66301 59.77425 61.34075 61.33823 60.97808 60.90758 60.90507
#>  [81] 60.90302 60.90103 60.89910 60.89724 60.89543 60.89367 60.89197 60.88396
#>  [89] 62.76372 62.76714 62.77046 62.77368 62.77743 62.82129 62.82426 62.82713
#>  [97] 62.82991 62.83259 62.83518 62.83769
#> 
#> $cvse
#>   [1] 10.000122  9.642985  8.820756  8.076346  7.390351  6.754897  6.171110
#>   [8]  5.647206  5.196858  4.837195  4.585608  4.455115  4.449555  4.561336
#>  [15]  4.773438  5.049952  5.014462  4.878662  4.782959  4.700885  4.629911
#>  [22]  4.568017  4.513598  4.452168  4.288149  4.194006  4.162633  4.249699
#>  [29]  4.238529  4.234046  4.533494  4.528051  4.834875  4.860499  4.857853
#>  [36]  4.855258  4.851878  4.885244  4.882609  4.880289  4.878136  5.699917
#>  [43]  5.696866  5.593523  5.812914 11.427686 12.234396 12.235824 12.643323
#>  [50] 12.656428 13.030991 13.032240 15.136424 18.130048 25.230804 25.195898
#>  [57] 25.151307 25.121838 25.294738 25.138132 25.147648 25.168912 28.151674
#>  [64] 28.179039 28.181905 28.165571 28.213924 28.218514 28.215691 28.218006
#>  [71] 28.220219 28.222335 26.641061 26.639696 26.625246 27.278064 27.276932
#>  [78] 27.305110 27.309580 27.308608 27.307645 27.306723 27.305842 27.304998
#>  [85] 27.304192 27.303421 27.302684 27.302382 28.903623 28.907220 28.910717
#>  [92] 28.914115 28.917400 28.919213 28.922320 28.925336 28.928260 28.931096
#>  [99] 28.933845 28.936508
#> 
#> $fold
#>  [1]  1  9  3  3  4  2 10  1  6  2 10  4  5  1  8  7  6  7  9  8  5
#> 
#> $lambda
#>   [1] 5.266887679 4.911917077 4.580870305 4.272134979 3.984207379 3.715685137
#>   [7] 3.465260394 3.231713442 3.013906772 2.810779543 2.621342409 2.444672704
#>  [13] 2.279909946 2.126251646 1.982949401 1.849305248 1.724668264 1.608431395
#>  [19] 1.500028502 1.398931602 1.304648294 1.216719365 1.134716552 1.058240454
#>  [25] 0.986918590 0.920403580 0.858371459 0.800520096 0.746567720 0.696251555
#>  [31] 0.649326530 0.605564095 0.564751102 0.526688768 0.491191708 0.458087031
#>  [37] 0.427213498 0.398420738 0.371568514 0.346526039 0.323171344 0.301390677
#>  [43] 0.281077954 0.262134241 0.244467270 0.227990995 0.212625165 0.198294940
#>  [49] 0.184930525 0.172466826 0.160843140 0.150002850 0.139893160 0.130464829
#>  [55] 0.121671937 0.113471655 0.105824045 0.098691859 0.092040358 0.085837146
#>  [61] 0.080052010 0.074656772 0.069625155 0.064932653 0.060556410 0.056475110
#>  [67] 0.052668877 0.049119171 0.045808703 0.042721350 0.039842074 0.037156851
#>  [73] 0.034652604 0.032317134 0.030139068 0.028107795 0.026213424 0.024446727
#>  [79] 0.022799099 0.021262516 0.019829494 0.018493052 0.017246683 0.016084314
#>  [85] 0.015000285 0.013989316 0.013046483 0.012167194 0.011347166 0.010582405
#>  [91] 0.009869186 0.009204036 0.008583715 0.008005201 0.007465677 0.006962516
#>  [97] 0.006493265 0.006055641 0.005647511 0.005266888
#> 
#> $fit
#> $beta
#>               5.2669     4.9119     4.5809     4.2721   3.9842    3.7157
#> (Intercept) 21.38095 23.1063067 24.7153779 26.2160031 27.61549 28.920658
#> am           0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#> cyl          0.00000 -0.2875591 -0.5557376 -0.8058418 -1.03909 -1.256618
#> disp         0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#> wt           0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#>                3.4653   3.2317    3.0139    2.8108    2.6213    2.4447
#> (Intercept) 30.137862 31.27303 32.331691 33.319002 34.239772 35.098485
#> am           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.459485 -1.64868 -1.825123 -1.989675 -2.143137 -2.286255
#> disp         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#>                2.2799    2.1263    1.9829    1.8493    1.7247    1.6084
#> (Intercept) 35.899324 36.646188 37.342717 37.992302 38.447619 38.447619
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.419729 -2.544206 -2.660294 -2.768558 -2.844444 -2.844444
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.5000    1.3989    1.3046    1.2167    1.1347    1.0582
#> (Intercept) 38.447619 38.447619 38.447619 38.447619 38.447619 38.447619
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.844444 -2.844444 -2.844444 -2.844444 -2.844444 -2.844444
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                 0.9869    0.9204    0.8584    0.8005   0.7466      0.6963
#> (Intercept) 38.9574997 40.191111 40.540301 40.540108 40.53999 40.53302779
#> am           0.0000000  0.000000  0.000000  0.000000  0.00000  0.00000000
#> carb         0.0000000  0.000000  0.000000  0.000000  0.00000 -0.01870702
#> cyl         -2.5639924 -1.882890 -1.689700 -1.689528 -1.68942 -1.67940798
#> disp         0.0000000  0.000000  0.000000  0.000000  0.00000  0.00000000
#> drat         0.0000000  0.000000  0.000000  0.000000  0.00000  0.00000000
#> gear         0.0000000  0.000000  0.000000  0.000000  0.00000  0.00000000
#> hp           0.0000000  0.000000  0.000000  0.000000  0.00000  0.00000000
#> qsec         0.0000000  0.000000  0.000000  0.000000  0.00000  0.00000000
#> vs           0.0000000  0.000000  0.000000  0.000000  0.00000  0.00000000
#> wt          -0.7186248 -2.462336 -2.956692 -2.956968 -2.95714 -2.95732156
#>                 0.6493     0.6056     0.5648     0.5267     0.4912     0.4581
#> (Intercept) 40.4962866 40.4620364 40.4300947 40.4003750 40.3725894 40.3466761
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb        -0.1203918 -0.2152069 -0.3036316 -0.3858929 -0.4628129 -0.5345497
#> cyl         -1.6259631 -1.5761367 -1.5296687 -1.4864356 -1.4460136 -1.4083154
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.9567455 -2.9561957 -2.9556829 -2.9552122 -2.9547658 -2.9543493
#>                 0.4272     0.3984     0.3716     0.3465     0.3232     0.3014
#> (Intercept) 40.3225094 40.2999714 40.2831795 40.2831516 40.2831435 40.2831420
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb        -0.6014517 -0.6638447 -0.7103057 -0.7104014 -0.7104358 -0.7104478
#> cyl         -1.3731580 -1.3403701 -1.3159466 -1.3159023 -1.3158881 -1.3158843
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.9539609 -2.9535987 -2.9533417 -2.9533316 -2.9533252 -2.9533211
#>                 0.2811     0.2621     0.2445     0.2280     0.2126     0.1983
#> (Intercept) 40.2831424 40.2831431 40.2831437 40.2831442 40.2831445 40.2831447
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb        -0.7104516 -0.7104527 -0.7104528 -0.7104527 -0.7104525 -0.7104524
#> cyl         -1.3158838 -1.3158843 -1.3158848 -1.3158853 -1.3158856 -1.3158858
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.9533186 -2.9533170 -2.9533160 -2.9533154 -2.9533150 -2.9533148
#>                 0.1849     0.1725     0.1608     0.1500     0.1399     0.1305
#> (Intercept) 40.2831449 40.2831450 40.2831450 21.2897289 21.2845437 21.2799874
#> am           0.0000000  0.0000000  0.0000000  2.1836848  2.1843541  2.1849370
#> carb        -0.7104523 -0.7104523 -0.7104522 -0.6877991 -0.6878370 -0.6878696
#> cyl         -1.3158859 -1.3158860 -1.3158861 -0.5065929 -0.5063185 -0.5060761
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  0.7318018  0.7319924  0.7321603
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.9533146 -2.9533146 -2.9533145 -3.0037035 -3.0037404 -3.0037776
#>                 0.1217     0.1135      0.1058      0.0987      0.0920
#> (Intercept) 21.2759794 21.2724500 14.42786183 14.43447010 10.03125575
#> am           2.1854459  2.1858908  3.00692381  3.00622345  5.12234291
#> carb        -0.6878976 -0.6879219  0.00000000  0.00000000  0.53654403
#> cyl         -0.5058618 -0.5056722 -0.63741107 -0.63762581  0.00000000
#> disp         0.0000000  0.0000000  0.02505041  0.02504610  0.03726721
#> drat         0.0000000  0.0000000  0.00000000  0.00000000 -1.78122668
#> gear         0.0000000  0.0000000  0.00000000  0.00000000  0.00000000
#> hp           0.0000000  0.0000000 -0.02214154 -0.02214189 -0.03930238
#> qsec         0.7323082  0.7324388  1.14469498  1.14440311  1.39594674
#> vs           0.0000000  0.0000000  0.00000000  0.00000000  3.06439923
#> wt          -3.0038139 -3.0038489 -4.37458762 -4.37418816 -4.87190011
#>                  0.0858      0.0801      0.0747      0.0696      0.0649
#> (Intercept) 10.02999235 10.02877829 10.02761165 10.02649059 10.02541332
#> am           5.12281307  5.12326485  5.12369899  5.12411616  5.12451703
#> carb         0.53669572  0.53684149  0.53698155  0.53711614  0.53724547
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.03727222  0.03727703  0.03728165  0.03728609  0.03729036
#> drat        -1.78163780 -1.78203285 -1.78241247 -1.78277725 -1.78312778
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.03930554 -0.03930858 -0.03931150 -0.03931430 -0.03931700
#> qsec         1.39607845  1.39620501  1.39632663  1.39644350  1.39655580
#> vs           3.06466808  3.06492642  3.06517467  3.06541320  3.06564241
#> wt          -4.87221784 -4.87252316 -4.87281655 -4.87309846 -4.87336937
#>                  0.0606      0.0565      0.0527      0.0491      0.0458
#> (Intercept) 10.02437814 10.02338340 10.95821995 10.95689185 10.95567057
#> am           5.12490224  5.12527239  5.38473730  5.38553134  5.38626208
#> carb         0.53736974  0.53748916  0.64708123  0.64725238  0.64740911
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.03729446  0.03729840  0.03832370  0.03832819  0.03833231
#> drat        -1.78346460 -1.78378827 -1.78367143 -1.78404821 -1.78439468
#> gear         0.00000000  0.00000000 -0.43196596 -0.43242015 -0.43283769
#> hp          -0.03931959 -0.03932207 -0.03882559 -0.03882376 -0.03882206
#> qsec         1.39666371  1.39676740  1.41342305  1.41362805  1.41381643
#> vs           3.06586266  3.06607429  3.43734319  3.43776932  3.43816062
#> wt          -4.87362968 -4.87387982 -5.03577352 -5.03627068 -5.03672639
#>                  0.0427      0.0398      0.0372      0.0347      0.0323
#> (Intercept) 10.95454841 10.95351803 10.95257249 10.95170525 10.95091020
#> am           5.38693427  5.38755235  5.38812045  5.38864238  5.38912169
#> carb         0.64755260  0.64768393  0.64780409  0.64791399  0.64801446
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.03833608  0.03833953  0.03834268  0.03834556  0.03834820
#> drat        -1.78471305 -1.78500540 -1.78527365 -1.78551965 -1.78574507
#> gear        -0.43322160 -0.43357462 -0.43389925 -0.43419777 -0.43447226
#> hp          -0.03882049 -0.03881902 -0.03881765 -0.03881638 -0.03881519
#> qsec         1.41398949  1.41414842  1.41429430  1.41442818  1.41455099
#> vs           3.43851988  3.43884966  3.43915227  3.43942987  3.43968441
#> wt          -5.03714406 -5.03752681 -5.03787752 -5.03819879 -5.03849304
#>                  0.0301      0.0281     0.0262      0.0244      0.0228
#> (Intercept) 10.95018162 10.94951421 10.9489031 10.94834360 10.94783163
#> am           5.38956169  5.38996542  5.3903357  5.39067518  5.39098625
#> carb         0.64810627  0.64819011  0.6482666  0.64833645  0.64840009
#> cyl          0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> disp         0.03835060  0.03835279  0.0383548  0.03835662  0.03835828
#> drat        -1.78595152 -1.78614045 -1.7863133 -1.78647118 -1.78661543
#> gear        -0.43472463 -0.43495663 -0.4351699 -0.43536584 -0.43554588
#> hp          -0.03881409 -0.03881306 -0.0388121 -0.03881120 -0.03881036
#> qsec         1.41466361  1.41476686  1.4148615  1.41494816  1.41502755
#> vs           3.43991771  3.44013142  3.4403271  3.44050614  3.44066988
#> wt          -5.03876248 -5.03900913 -5.0392348 -5.03944134 -5.03963018
#>                  0.0213      0.0198       0.0185      0.0172      0.0161
#> (Intercept) 10.94736327 10.94693493 10.944757978  7.21242618  7.21118130
#> am           5.39127117  5.39153200  5.391770671  5.85382954  5.85443042
#> carb         0.64845807  0.64851084  0.648558838  0.70562587  0.70581489
#> cyl          0.00000000  0.00000000  0.000299246  0.34695613  0.34701033
#> disp         0.03835980  0.03836117  0.038358022  0.03940763  0.03941272
#> drat        -1.78674708 -1.78686714 -1.786899322 -1.94267095 -1.94306005
#> gear        -0.43571126 -0.43586313 -0.436016074 -0.34848342 -0.34865149
#> hp          -0.03880958 -0.03880885 -0.038808432 -0.04262590 -0.04262914
#> qsec         1.41510022  1.41516673  1.415255099  1.50962861  1.50976190
#> vs           3.44081951  3.44095615  3.441110256  4.09700883  4.09753048
#> wt          -5.03980281 -5.03996056 -5.040035803 -5.09948460 -5.09983928
#>                  0.0150      0.0140      0.0130      0.0122      0.0113
#> (Intercept)  7.20998578  7.20883799  7.20773636  7.20667934  7.20566546
#> am           5.85501316  5.85557821  5.85612604  5.85665708  5.85717176
#> carb         0.70599856  0.70617703  0.70635040  0.70651880  0.70668235
#> cyl          0.34706153  0.34710983  0.34715537  0.34719824  0.34723856
#> disp         0.03941767  0.03942248  0.03942715  0.03943169  0.03943610
#> drat        -1.94343792 -1.94380480 -1.94416098 -1.94450672 -1.94484226
#> gear        -0.34881565 -0.34897597 -0.34913251 -0.34928533 -0.34943450
#> hp          -0.04263228 -0.04263531 -0.04263824 -0.04264107 -0.04264381
#> qsec         1.50989128  1.51001683  1.51013864  1.51025682  1.51037146
#> vs           4.09803495  4.09852272  4.09899426  4.09945003  4.09989048
#> wt          -5.10018430 -5.10051986 -5.10084617 -5.10116345 -5.10147189
#>                  0.0106      0.0099      0.0092      0.0086      0.0080
#> (Intercept)  7.20469326  7.20376133  7.20286831  7.20201287  7.20119372
#> am           5.85767053  5.85815379  5.85862194  5.85907540  5.85951455
#> carb         0.70684116  0.70699534  0.70714501  0.70729029  0.70743127
#> cyl          0.34727644  0.34731196  0.34734524  0.34737637  0.34740542
#> disp         0.03944038  0.03944454  0.03944858  0.03945250  0.03945630
#> drat        -1.94516786 -1.94548377 -1.94579022 -1.94608747 -1.94637573
#> gear        -0.34958007 -0.34972212 -0.34986070 -0.34999587 -0.35012770
#> hp          -0.04264645 -0.04264900 -0.04265146 -0.04265383 -0.04265612
#> qsec         1.51048264  1.51059045  1.51069497  1.51079630  1.51089451
#> vs           4.10031605  4.10072717  4.10112425  4.10150770  4.10187792
#> wt          -5.10177170 -5.10206308 -5.10234622 -5.10262131 -5.10288855
#>                  0.0075      0.0070      0.0065      0.0061      0.0056
#> (Intercept)  7.20040961  7.19965932  7.19894168  7.19825555  7.19759980
#> am           5.85993977  5.86035145  5.86074994  5.86113561  5.86150881
#> carb         0.70756806  0.70770077  0.70782950  0.70795435  0.70807542
#> cyl          0.34743251  0.34745770  0.34748108  0.34750273  0.34752272
#> disp         0.03945999  0.03946358  0.03946705  0.03947042  0.03947369
#> drat        -1.94665524 -1.94692623 -1.94718890 -1.94744349 -1.94769019
#> gear        -0.35025626 -0.35038159 -0.35050376 -0.35062284 -0.35073888
#> hp          -0.04265834 -0.04266047 -0.04266252 -0.04266451 -0.04266642
#> qsec         1.51098969  1.51108191  1.51117125  1.51125779  1.51134161
#> vs           4.10223531  4.10258023  4.10291305  4.10323415  4.10354386
#> wt          -5.10314811 -5.10340019 -5.10364495 -5.10388258 -5.10411325
#>                  0.0053
#> (Intercept)  7.19697337
#> am           5.86186987
#> carb         0.70819281
#> cyl          0.34754113
#> disp         0.03947686
#> drat        -1.94792921
#> gear        -0.35085194
#> hp          -0.04266826
#> qsec         1.51142277
#> vs           4.10384253
#> wt          -5.10433713
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1   1  67  80  15   1   1   8  10  10  10   9   9   9
#>  [37]   9   9   6   1   1   1   1   1   1   1   1   1   1   1   1 191   1   1
#>  [55]   1   1 148   1 266   1   1   1   1   1   1   1  46   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1 153   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.266887679 4.911917077 4.580870305 4.272134979 3.984207379 3.715685137
#>   [7] 3.465260394 3.231713442 3.013906772 2.810779543 2.621342409 2.444672704
#>  [13] 2.279909946 2.126251646 1.982949401 1.849305248 1.724668264 1.608431395
#>  [19] 1.500028502 1.398931602 1.304648294 1.216719365 1.134716552 1.058240454
#>  [25] 0.986918590 0.920403580 0.858371459 0.800520096 0.746567720 0.696251555
#>  [31] 0.649326530 0.605564095 0.564751102 0.526688768 0.491191708 0.458087031
#>  [37] 0.427213498 0.398420738 0.371568514 0.346526039 0.323171344 0.301390677
#>  [43] 0.281077954 0.262134241 0.244467270 0.227990995 0.212625165 0.198294940
#>  [49] 0.184930525 0.172466826 0.160843140 0.150002850 0.139893160 0.130464829
#>  [55] 0.121671937 0.113471655 0.105824045 0.098691859 0.092040358 0.085837146
#>  [61] 0.080052010 0.074656772 0.069625155 0.064932653 0.060556410 0.056475110
#>  [67] 0.052668877 0.049119171 0.045808703 0.042721350 0.039842074 0.037156851
#>  [73] 0.034652604 0.032317134 0.030139068 0.028107795 0.026213424 0.024446727
#>  [79] 0.022799099 0.021262516 0.019829494 0.018493052 0.017246683 0.016084314
#>  [85] 0.015000285 0.013989316 0.013046483 0.012167194 0.011347166 0.010582405
#>  [91] 0.009869186 0.009204036 0.008583715 0.008005201 0.007465677 0.006962516
#>  [97] 0.006493265 0.006055641 0.005647511 0.005266888
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 24
#> 
#> $loss
#>   [1] 787.7724 675.9419 582.3791 504.4550 439.9001 386.7561 343.3343 308.1796
#>   [9] 280.0393 257.8356 240.6421 227.6636 218.2178 211.7205 207.6718 205.6448
#>  [17] 205.2302 205.2302 205.2302 205.2302 205.2302 205.2302 205.2302 205.2302
#>  [25] 180.6153 149.1912 147.5754 147.5754 147.5754 146.7050 142.3800 138.9654
#>  [33] 136.3186 134.3223 132.8618 131.8536 131.2211 130.8990 130.8269 130.8269
#>  [41] 130.8269 130.8269 130.8269 130.8269 130.8269 130.8269 130.8269 130.8269
#>  [49] 130.8269 130.8269 130.8269 123.6839 123.6839 123.6839 123.6839 123.6839
#>  [57] 115.1112 115.1112 110.1067 110.1067 110.1067 110.1067 110.1067 110.1067
#>  [65] 110.1066 110.1066 109.7625 109.7625 109.7625 109.7625 109.7625 109.7625
#>  [73] 109.7625 109.7625 109.7624 109.7624 109.7624 109.7624 109.7624 109.7624
#>  [81] 109.7624 109.7620 109.5097 109.5097 109.5097 109.5097 109.5096 109.5096
#>  [89] 109.5096 109.5096 109.5096 109.5096 109.5096 109.5096 109.5096 109.5095
#>  [97] 109.5095 109.5095 109.5095 109.5095
#> 
#> $linear.predictors
#>      5.2669   4.9119   4.5809   4.2721   3.9842   3.7157   3.4653   3.2317
#> 1  21.38095 21.38095 21.38095 21.38095 21.38095 21.38095 21.38095 21.38095
#> 2  21.38095 21.38095 21.38095 21.38095 21.38095 21.38095 21.38095 21.38095
#> 3  21.38095 21.95607 22.49243 22.99264 23.45913 23.89419 24.29992 24.67831
#> 4  21.38095 20.80583 20.26948 19.76927 19.30277 18.86772 18.46198 18.08359
#> 5  21.38095 21.95607 22.49243 22.99264 23.45913 23.89419 24.29992 24.67831
#> 6  21.38095 21.95607 22.49243 22.99264 23.45913 23.89419 24.29992 24.67831
#> 7  21.38095 21.38095 21.38095 21.38095 21.38095 21.38095 21.38095 21.38095
#> 8  21.38095 20.80583 20.26948 19.76927 19.30277 18.86772 18.46198 18.08359
#> 9  21.38095 20.80583 20.26948 19.76927 19.30277 18.86772 18.46198 18.08359
#> 10 21.38095 20.80583 20.26948 19.76927 19.30277 18.86772 18.46198 18.08359
#> 11 21.38095 20.80583 20.26948 19.76927 19.30277 18.86772 18.46198 18.08359
#> 12 21.38095 21.95607 22.49243 22.99264 23.45913 23.89419 24.29992 24.67831
#> 13 21.38095 21.95607 22.49243 22.99264 23.45913 23.89419 24.29992 24.67831
#> 14 21.38095 21.95607 22.49243 22.99264 23.45913 23.89419 24.29992 24.67831
#> 15 21.38095 20.80583 20.26948 19.76927 19.30277 18.86772 18.46198 18.08359
#> 16 21.38095 20.80583 20.26948 19.76927 19.30277 18.86772 18.46198 18.08359
#> 17 21.38095 21.95607 22.49243 22.99264 23.45913 23.89419 24.29992 24.67831
#> 18 21.38095 21.95607 22.49243 22.99264 23.45913 23.89419 24.29992 24.67831
#> 19 21.38095 20.80583 20.26948 19.76927 19.30277 18.86772 18.46198 18.08359
#> 20 21.38095 20.80583 20.26948 19.76927 19.30277 18.86772 18.46198 18.08359
#> 21 21.38095 21.95607 22.49243 22.99264 23.45913 23.89419 24.29992 24.67831
#>      3.0139   2.8108   2.6213   2.4447   2.2799   2.1263   1.9829   1.8493
#> 1  21.38095 21.38095 21.38095 21.38095 21.38095 21.38095 21.38095 21.38095
#> 2  21.38095 21.38095 21.38095 21.38095 21.38095 21.38095 21.38095 21.38095
#> 3  25.03120 25.36030 25.66723 25.95346 26.22041 26.46936 26.70154 26.91807
#> 4  17.73071 17.40160 17.09468 16.80844 16.54150 16.29254 16.06036 15.84384
#> 5  25.03120 25.36030 25.66723 25.95346 26.22041 26.46936 26.70154 26.91807
#> 6  25.03120 25.36030 25.66723 25.95346 26.22041 26.46936 26.70154 26.91807
#> 7  21.38095 21.38095 21.38095 21.38095 21.38095 21.38095 21.38095 21.38095
#> 8  17.73071 17.40160 17.09468 16.80844 16.54150 16.29254 16.06036 15.84384
#> 9  17.73071 17.40160 17.09468 16.80844 16.54150 16.29254 16.06036 15.84384
#> 10 17.73071 17.40160 17.09468 16.80844 16.54150 16.29254 16.06036 15.84384
#> 11 17.73071 17.40160 17.09468 16.80844 16.54150 16.29254 16.06036 15.84384
#> 12 25.03120 25.36030 25.66723 25.95346 26.22041 26.46936 26.70154 26.91807
#> 13 25.03120 25.36030 25.66723 25.95346 26.22041 26.46936 26.70154 26.91807
#> 14 25.03120 25.36030 25.66723 25.95346 26.22041 26.46936 26.70154 26.91807
#> 15 17.73071 17.40160 17.09468 16.80844 16.54150 16.29254 16.06036 15.84384
#> 16 17.73071 17.40160 17.09468 16.80844 16.54150 16.29254 16.06036 15.84384
#> 17 25.03120 25.36030 25.66723 25.95346 26.22041 26.46936 26.70154 26.91807
#> 18 25.03120 25.36030 25.66723 25.95346 26.22041 26.46936 26.70154 26.91807
#> 19 17.73071 17.40160 17.09468 16.80844 16.54150 16.29254 16.06036 15.84384
#> 20 17.73071 17.40160 17.09468 16.80844 16.54150 16.29254 16.06036 15.84384
#> 21 25.03120 25.36030 25.66723 25.95346 26.22041 26.46936 26.70154 26.91807
#>      1.7247   1.6084   1.5000   1.3989   1.3046   1.2167   1.1347   1.0582
#> 1  21.38095 21.38095 21.38095 21.38095 21.38095 21.38095 21.38095 21.38095
#> 2  21.38095 21.38095 21.38095 21.38095 21.38095 21.38095 21.38095 21.38095
#> 3  27.06984 27.06984 27.06984 27.06984 27.06984 27.06984 27.06984 27.06984
#> 4  15.69206 15.69206 15.69206 15.69206 15.69206 15.69206 15.69206 15.69206
#> 5  27.06984 27.06984 27.06984 27.06984 27.06984 27.06984 27.06984 27.06984
#> 6  27.06984 27.06984 27.06984 27.06984 27.06984 27.06984 27.06984 27.06984
#> 7  21.38095 21.38095 21.38095 21.38095 21.38095 21.38095 21.38095 21.38095
#> 8  15.69206 15.69206 15.69206 15.69206 15.69206 15.69206 15.69206 15.69206
#> 9  15.69206 15.69206 15.69206 15.69206 15.69206 15.69206 15.69206 15.69206
#> 10 15.69206 15.69206 15.69206 15.69206 15.69206 15.69206 15.69206 15.69206
#> 11 15.69206 15.69206 15.69206 15.69206 15.69206 15.69206 15.69206 15.69206
#> 12 27.06984 27.06984 27.06984 27.06984 27.06984 27.06984 27.06984 27.06984
#> 13 27.06984 27.06984 27.06984 27.06984 27.06984 27.06984 27.06984 27.06984
#> 14 27.06984 27.06984 27.06984 27.06984 27.06984 27.06984 27.06984 27.06984
#> 15 15.69206 15.69206 15.69206 15.69206 15.69206 15.69206 15.69206 15.69206
#> 16 15.69206 15.69206 15.69206 15.69206 15.69206 15.69206 15.69206 15.69206
#> 17 27.06984 27.06984 27.06984 27.06984 27.06984 27.06984 27.06984 27.06984
#> 18 27.06984 27.06984 27.06984 27.06984 27.06984 27.06984 27.06984 27.06984
#> 19 15.69206 15.69206 15.69206 15.69206 15.69206 15.69206 15.69206 15.69206
#> 20 15.69206 15.69206 15.69206 15.69206 15.69206 15.69206 15.69206 15.69206
#> 21 27.06984 27.06984 27.06984 27.06984 27.06984 27.06984 27.06984 27.06984
#>      0.9869   0.9204   0.8584   0.8005   0.7466   0.6963   0.6493   0.6056
#> 1  21.69075 22.44245 22.65557 22.65569 22.65576 22.63357 22.51227 22.39916
#> 2  21.50750 21.81456 21.90161 21.90166 21.90169 21.87945 21.75830 21.64533
#> 3  27.03432 26.94693 26.92198 26.92183 26.92174 26.93570 27.01239 27.08391
#> 4  15.88007 16.33746 16.46731 16.46751 16.46764 16.46530 16.45143 16.43850
#> 5  26.40912 24.80470 24.34965 24.34927 24.34903 24.34413 24.31963 24.29681
#> 6  26.43786 24.90319 24.46792 24.46755 24.46732 24.46242 24.43790 24.41506
#> 7  21.10148 20.42334 20.23108 20.23097 20.23091 20.20857 20.08774 19.97508
#> 8  15.52076 15.10629 14.98896 14.98903 14.98907 15.00534 15.09345 15.17561
#> 9  15.76509 15.94348 15.99424 15.99440 15.99450 16.01083 16.09875 16.18071
#> 10 15.72916 15.82036 15.84641 15.84655 15.84664 15.86297 15.95091 16.03290
#> 11 14.60451 11.96681 11.21918 11.21889 11.21871 11.21605 11.20321 11.19125
#> 12 27.12056 27.24241 27.27678 27.27667 27.27660 27.29058 27.36720 27.43865
#> 13 27.54095 28.68288 29.00644 29.00649 29.00653 29.00191 28.97651 28.95282
#> 14 27.38285 28.14117 28.35597 28.35596 28.35596 28.37000 28.44641 28.51766
#> 15 15.91600 16.46057 16.61515 16.61536 16.61550 16.65058 16.84005 17.01672
#> 16 15.68245 15.66031 15.65422 15.65435 15.65442 15.68945 15.87911 16.05596
#> 17 27.31099 27.89493 28.06030 28.06027 28.06024 28.07427 28.15074 28.22204
#> 18 27.61425 28.93404 29.30803 29.30811 29.30816 29.30355 29.27809 29.25435
#> 19 16.16752 17.32239 17.64999 17.65030 17.65049 17.64823 17.63413 17.62097
#> 20 15.88007 16.33746 16.46731 16.46751 16.46764 16.39047 15.96987 15.57767
#> 21 26.70375 25.81426 25.56190 25.56163 25.56146 25.55663 25.53190 25.50885
#>      0.5648   0.5267   0.4912   0.4581   0.4272   0.3984   0.3716   0.3465
#> 1  22.29367 22.19553 22.10377 22.01819 21.93838 21.86394 21.80852 21.80840
#> 2  21.53997 21.44195 21.35030 21.26483 21.18512 21.11078 21.05542 21.05530
#> 3  27.15060 27.21265 27.27067 27.32477 27.37524 27.42230 27.45733 27.45741
#> 4  16.42643 16.41521 16.40472 16.39493 16.38580 16.37728 16.37095 16.37093
#> 5  24.27553 24.25572 24.23721 24.21994 24.20384 24.18882 24.17762 24.17761
#> 6  24.39376 24.37393 24.35540 24.33811 24.32200 24.30697 24.29576 24.29575
#> 7  19.87001 19.77226 19.68086 19.59562 19.51613 19.44199 19.38678 19.38667
#> 8  15.25222 15.32350 15.39015 15.45230 15.51027 15.56433 15.60459 15.60467
#> 9  16.25715 16.32827 16.39477 16.45678 16.51462 16.56855 16.60873 16.60880
#> 10 16.10937 16.18051 16.24703 16.30906 16.36692 16.42087 16.46106 16.46114
#> 11 11.18009 11.16971 11.16001 11.15096 11.14252 11.13465 11.12877 11.12877
#> 12 27.50529 27.56727 27.62524 27.67930 27.72971 27.77673 27.81174 27.81181
#> 13 28.93073 28.91018 28.89096 28.87304 28.85633 28.84074 28.82913 28.82911
#> 14 28.58411 28.64593 28.70373 28.75763 28.80791 28.85479 28.88971 28.88978
#> 15 17.18148 17.33476 17.47808 17.61174 17.73640 17.85265 17.93923 17.93940
#> 16 16.22088 16.37431 16.51778 16.65158 16.77636 16.89273 16.97940 16.97957
#> 17 28.28854 28.35040 28.40825 28.46220 28.51251 28.55943 28.59437 28.59444
#> 18 29.23221 29.21161 29.19235 29.17438 29.15763 29.14201 29.13038 29.13035
#> 19 17.60870 17.59730 17.58662 17.57667 17.56738 17.55872 17.55229 17.55227
#> 20 15.21190 14.87164 14.55346 14.25673 13.97999 13.72191 13.52973 13.52933
#> 21 25.48736 25.46736 25.44866 25.43122 25.41496 25.39980 25.38849 25.38848
#>      0.3232   0.3014   0.2811   0.2621   0.2445   0.2280   0.2126   0.1983
#> 1  21.80836 21.80834 21.80834 21.80834 21.80834 21.80834 21.80834 21.80834
#> 2  21.05526 21.05525 21.05524 21.05524 21.05524 21.05524 21.05524 21.05524
#> 3  27.45744 27.45745 27.45746 27.45746 27.45746 27.45746 27.45746 27.45746
#> 4  16.37092 16.37092 16.37092 16.37092 16.37092 16.37092 16.37092 16.37092
#> 5  24.17761 24.17761 24.17762 24.17762 24.17762 24.17762 24.17762 24.17762
#> 6  24.29575 24.29575 24.29575 24.29575 24.29575 24.29575 24.29575 24.29576
#> 7  19.38663 19.38662 19.38662 19.38662 19.38662 19.38662 19.38662 19.38662
#> 8  15.60470 15.60471 15.60471 15.60471 15.60471 15.60471 15.60471 15.60471
#> 9  16.60883 16.60884 16.60884 16.60884 16.60884 16.60884 16.60884 16.60884
#> 10 16.46116 16.46117 16.46117 16.46117 16.46117 16.46117 16.46117 16.46117
#> 11 11.12877 11.12878 11.12878 11.12878 11.12878 11.12878 11.12878 11.12878
#> 12 27.81184 27.81185 27.81185 27.81186 27.81186 27.81186 27.81186 27.81186
#> 13 28.82910 28.82910 28.82909 28.82909 28.82909 28.82909 28.82909 28.82909
#> 14 28.88980 28.88981 28.88982 28.88982 28.88982 28.88982 28.88982 28.88982
#> 15 17.93946 17.93948 17.93949 17.93949 17.93949 17.93949 17.93949 17.93949
#> 16 16.97963 16.97965 16.97966 16.97966 16.97966 16.97966 16.97966 16.97966
#> 17 28.59447 28.59448 28.59448 28.59448 28.59449 28.59449 28.59449 28.59449
#> 18 29.13034 29.13033 29.13033 29.13033 29.13033 29.13033 29.13033 29.13033
#> 19 17.55225 17.55225 17.55225 17.55224 17.55224 17.55224 17.55224 17.55224
#> 20 13.52918 13.52913 13.52911 13.52911 13.52910 13.52910 13.52911 13.52911
#> 21 25.38848 25.38848 25.38848 25.38848 25.38848 25.38848 25.38848 25.38848
#>      0.1849   0.1725   0.1608   0.1500   0.1399   0.1305   0.1217   0.1135
#> 1  21.80834 21.80834 21.80834 21.85841 21.85843 21.85845 21.85846 21.85848
#> 2  21.05524 21.05524 21.05524 21.50228 21.50240 21.50250 21.50258 21.50266
#> 3  27.45746 27.45746 27.45746 27.40948 27.40949 27.40949 27.40949 27.40949
#> 4  16.37091 16.37091 16.37091 15.35431 15.35405 15.35383 15.35364 15.35347
#> 5  24.17762 24.17762 24.17762 22.94198 22.94151 22.94110 22.94074 22.94041
#> 6  24.29576 24.29576 24.29576 25.18435 25.18444 25.18451 25.18458 25.18464
#> 7  19.38662 19.38662 19.38662 18.99729 18.99707 18.99689 18.99672 18.99658
#> 8  15.60471 15.60471 15.60471 15.68187 15.68193 15.68198 15.68203 15.68208
#> 9  16.60884 16.60884 16.60884 16.84949 16.84960 16.84970 16.84979 16.84987
#> 10 16.46117 16.46117 16.46117 16.99202 16.99221 16.99238 16.99252 16.99266
#> 11 11.12878 11.12878 11.12878 11.17898 11.17896 11.17894 11.17892 11.17890
#> 12 27.81186 27.81186 27.81186 28.39928 28.39945 28.39960 28.39973 28.39984
#> 13 28.82909 28.82909 28.82909 28.77343 28.77341 28.77339 28.77337 28.77336
#> 14 28.88982 28.88982 28.88982 29.81030 29.81057 29.81081 29.81102 29.81120
#> 15 17.93949 17.93949 17.93948 17.63385 17.63387 17.63389 17.63390 17.63392
#> 16 16.97966 16.97966 16.97966 16.78937 16.78941 16.78945 16.78948 16.78951
#> 17 28.59449 28.59448 28.59448 28.77813 28.77821 28.77827 28.77833 28.77837
#> 18 29.13033 29.13033 29.13033 27.89429 27.89396 27.89367 27.89342 27.89320
#> 19 17.55224 17.55224 17.55224 17.75886 17.75903 17.75919 17.75932 17.75944
#> 20 13.52911 13.52911 13.52911 13.87936 13.87939 13.87941 13.87943 13.87945
#> 21 25.38848 25.38848 25.38848 25.33266 25.33261 25.33256 25.33251 25.33247
#>      0.1058   0.0987   0.0920   0.0858   0.0801   0.0747   0.0696   0.0649
#> 1  22.56308 22.56321 22.20539 22.20539 22.20539 22.20538 22.20538 22.20538
#> 2  22.08858 22.08866 21.74478 21.74477 21.74477 21.74476 21.74475 21.74474
#> 3  26.68515 26.68520 26.94232 26.94232 26.94233 26.94233 26.94234 26.94234
#> 4  15.43673 15.43679 14.96592 14.96592 14.96593 14.96593 14.96593 14.96594
#> 5  23.11930 23.11983 23.00394 23.00389 23.00385 23.00380 23.00376 23.00371
#> 6  25.73543 25.73512 25.32053 25.32064 25.32076 25.32086 25.32097 25.32107
#> 7  18.66459 18.66500 19.29527 19.29512 19.29497 19.29483 19.29469 19.29456
#> 8  14.36512 14.36531 13.83723 13.83697 13.83672 13.83648 13.83625 13.83603
#> 9  16.08142 16.08141 15.77287 15.77274 15.77262 15.77250 15.77239 15.77228
#> 10 16.32057 16.32046 16.08765 16.08756 16.08747 16.08739 16.08731 16.08724
#> 11 11.81661 11.81658 12.05918 12.05927 12.05935 12.05943 12.05951 12.05958
#> 12 28.05839 28.05827 28.28701 28.28701 28.28701 28.28701 28.28701 28.28702
#> 13 29.76489 29.76483 29.27186 29.27175 29.27165 29.27155 29.27145 29.27136
#> 14 29.97909 29.97873 30.17221 30.17229 30.17237 30.17245 30.17252 30.17259
#> 15 17.88583 17.88578 18.54431 18.54443 18.54456 18.54467 18.54478 18.54489
#> 16 18.17073 18.17039 18.71557 18.71582 18.71605 18.71628 18.71650 18.71671
#> 17 28.57269 28.57263 28.79356 28.79357 28.79358 28.79359 28.79360 28.79361
#> 18 27.49204 27.49231 27.89912 27.89921 27.89930 27.89939 27.89947 27.89955
#> 19 18.01346 18.01308 17.28527 17.28517 17.28508 17.28499 17.28490 17.28482
#> 20 13.55352 13.55346 14.17968 14.17988 14.18008 14.18026 14.18044 14.18061
#> 21 24.63279 24.63296 24.61634 24.61626 24.61618 24.61610 24.61603 24.61596
#>      0.0606   0.0565   0.0527   0.0491   0.0458   0.0427   0.0398   0.0372
#> 1  22.20538 22.20538 22.17929 22.17915 22.17902 22.17889 22.17878 22.17868
#> 2  21.74474 21.74473 21.68669 21.68653 21.68639 21.68626 21.68614 21.68602
#> 3  26.94235 26.94235 26.98137 26.98148 26.98159 26.98169 26.98177 26.98186
#> 4  14.96594 14.96594 15.22023 15.22056 15.22085 15.22112 15.22137 15.22160
#> 5  23.00367 23.00364 22.79936 22.79888 22.79844 22.79803 22.79765 22.79731
#> 6  25.32116 25.32126 25.18212 25.18220 25.18227 25.18234 25.18240 25.18246
#> 7  19.29444 19.29432 19.30217 19.30180 19.30146 19.30115 19.30086 19.30060
#> 8  13.83581 13.83561 13.80673 13.80651 13.80630 13.80612 13.80595 13.80579
#> 9  15.77218 15.77208 15.80158 15.80156 15.80155 15.80154 15.80153 15.80153
#> 10 16.08716 16.08709 16.11516 16.11520 16.11524 16.11528 16.11532 16.11535
#> 11 12.05966 12.05972 12.12755 12.12764 12.12772 12.12780 12.12786 12.12793
#> 12 28.28702 28.28702 28.31637 28.31645 28.31653 28.31660 28.31666 28.31672
#> 13 29.27127 29.27119 29.47909 29.47908 29.47907 29.47906 29.47905 29.47904
#> 14 30.17265 30.17272 30.26005 30.26031 30.26055 30.26077 30.26098 30.26117
#> 15 18.54500 18.54510 18.51517 18.51520 18.51522 18.51524 18.51525 18.51527
#> 16 18.71691 18.71711 18.73409 18.73428 18.73446 18.73462 18.73477 18.73490
#> 17 28.79362 28.79363 28.85670 28.85679 28.85688 28.85697 28.85704 28.85711
#> 18 27.89963 27.89970 27.71521 27.71510 27.71500 27.71490 27.71481 27.71473
#> 19 17.28474 17.28467 16.97725 16.97700 16.97677 16.97655 16.97636 16.97617
#> 20 14.18078 14.18094 14.23271 14.23312 14.23350 14.23385 14.23417 14.23447
#> 21 24.61589 24.61583 24.71111 24.71115 24.71119 24.71123 24.71126 24.71130
#>      0.0347   0.0323   0.0301   0.0281   0.0262   0.0244   0.0228   0.0213
#> 1  22.17858 22.17849 22.17841 22.17834 22.17827 22.17820 22.17814 22.17809
#> 2  21.68592 21.68583 21.68574 21.68566 21.68558 21.68551 21.68545 21.68539
#> 3  26.98193 26.98201 26.98207 26.98213 26.98219 26.98224 26.98229 26.98233
#> 4  15.22181 15.22200 15.22218 15.22235 15.22249 15.22263 15.22276 15.22287
#> 5  22.79699 22.79670 22.79643 22.79618 22.79596 22.79575 22.79556 22.79538
#> 6  25.18251 25.18256 25.18260 25.18264 25.18268 25.18271 25.18274 25.18277
#> 7  19.30035 19.30013 19.29992 19.29973 19.29956 19.29940 19.29925 19.29912
#> 8  13.80565 13.80552 13.80540 13.80529 13.80519 13.80510 13.80502 13.80494
#> 9  15.80152 15.80152 15.80151 15.80151 15.80150 15.80150 15.80150 15.80150
#> 10 16.11538 16.11541 16.11544 16.11546 16.11549 16.11551 16.11553 16.11555
#> 11 12.12799 12.12804 12.12809 12.12813 12.12817 12.12821 12.12824 12.12828
#> 12 28.31677 28.31682 28.31687 28.31692 28.31696 28.31699 28.31703 28.31706
#> 13 29.47903 29.47902 29.47901 29.47900 29.47899 29.47899 29.47898 29.47897
#> 14 30.26134 30.26150 30.26164 30.26178 30.26190 30.26201 30.26211 30.26221
#> 15 18.51528 18.51529 18.51530 18.51531 18.51532 18.51533 18.51533 18.51534
#> 16 18.73502 18.73514 18.73524 18.73533 18.73541 18.73549 18.73556 18.73562
#> 17 28.85718 28.85724 28.85729 28.85734 28.85739 28.85743 28.85747 28.85751
#> 18 27.71466 27.71459 27.71453 27.71447 27.71442 27.71437 27.71433 27.71428
#> 19 16.97601 16.97586 16.97572 16.97559 16.97547 16.97536 16.97527 16.97518
#> 20 14.23474 14.23499 14.23522 14.23543 14.23562 14.23580 14.23597 14.23611
#> 21 24.71133 24.71136 24.71138 24.71141 24.71143 24.71146 24.71148 24.71150
#>      0.0198   0.0185   0.0172   0.0161   0.0150   0.0140   0.0130   0.0122
#> 1  22.17804 22.17815 22.10435 22.10432 22.10430 22.10427 22.10424 22.10422
#> 2  21.68534 21.68548 21.64938 21.64933 21.64929 21.64924 21.64920 21.64916
#> 3  26.98237 26.98226 26.93870 26.93872 26.93874 26.93875 26.93877 26.93879
#> 4  15.22298 15.22293 14.97991 14.97999 14.98006 14.98013 14.98021 14.98028
#> 5  22.79522 22.79485 22.60963 22.60948 22.60932 22.60918 22.60903 22.60890
#> 6  25.18280 25.18271 25.10556 25.10558 25.10560 25.10561 25.10563 25.10565
#> 7  19.29899 19.29916 19.55596 19.55587 19.55579 19.55570 19.55562 19.55554
#> 8  13.80488 13.80512 13.80410 13.80385 13.80361 13.80338 13.80315 13.80293
#> 9  15.80150 15.80178 15.83985 15.83975 15.83965 15.83956 15.83947 15.83938
#> 10 16.11556 16.11588 16.18873 16.18866 16.18860 16.18854 16.18848 16.18842
#> 11 12.12830 12.12800 12.06669 12.06679 12.06689 12.06699 12.06709 12.06718
#> 12 28.31709 28.31714 28.39836 28.39838 28.39841 28.39843 28.39845 28.39847
#> 13 29.47897 29.47901 29.48030 29.48030 29.48029 29.48029 29.48028 29.48028
#> 14 30.26230 30.26243 30.37996 30.38009 30.38020 30.38032 30.38043 30.38054
#> 15 18.51534 18.51539 18.64710 18.64722 18.64733 18.64745 18.64756 18.64766
#> 16 18.73568 18.73547 18.80562 18.80586 18.80610 18.80632 18.80655 18.80676
#> 17 28.85754 28.85756 28.90106 28.90110 28.90114 28.90119 28.90123 28.90126
#> 18 27.71425 27.71399 27.62420 27.62420 27.62419 27.62419 27.62419 27.62419
#> 19 16.97509 16.97489 17.02701 17.02681 17.02662 17.02643 17.02625 17.02607
#> 20 14.23625 14.23643 14.28488 14.28509 14.28529 14.28549 14.28568 14.28587
#> 21 24.71152 24.71138 24.60865 24.60861 24.60857 24.60853 24.60849 24.60845
#>      0.0113   0.0106   0.0099   0.0092   0.0086   0.0080   0.0075   0.0070
#> 1  22.10419 22.10417 22.10414 22.10412 22.10410 22.10408 22.10406 22.10404
#> 2  21.64912 21.64909 21.64905 21.64901 21.64898 21.64894 21.64891 21.64888
#> 3  26.93881 26.93882 26.93884 26.93886 26.93887 26.93889 26.93890 26.93892
#> 4  14.98035 14.98042 14.98049 14.98056 14.98062 14.98069 14.98075 14.98082
#> 5  22.60876 22.60863 22.60851 22.60839 22.60827 22.60815 22.60804 22.60794
#> 6  25.10567 25.10568 25.10570 25.10572 25.10574 25.10575 25.10577 25.10579
#> 7  19.55546 19.55538 19.55530 19.55522 19.55514 19.55507 19.55499 19.55492
#> 8  13.80272 13.80251 13.80230 13.80211 13.80191 13.80173 13.80155 13.80137
#> 9  15.83929 15.83921 15.83912 15.83904 15.83897 15.83889 15.83882 15.83874
#> 10 16.18837 16.18831 16.18826 16.18820 16.18815 16.18810 16.18805 16.18801
#> 11 12.06727 12.06736 12.06745 12.06753 12.06761 12.06769 12.06777 12.06785
#> 12 28.39849 28.39851 28.39853 28.39855 28.39857 28.39858 28.39860 28.39861
#> 13 29.48027 29.48027 29.48026 29.48026 29.48026 29.48025 29.48025 29.48024
#> 14 30.38064 30.38074 30.38084 30.38094 30.38103 30.38112 30.38120 30.38129
#> 15 18.64776 18.64786 18.64796 18.64805 18.64814 18.64823 18.64831 18.64839
#> 16 18.80697 18.80717 18.80737 18.80756 18.80774 18.80792 18.80810 18.80827
#> 17 28.90130 28.90134 28.90138 28.90141 28.90144 28.90148 28.90151 28.90154
#> 18 27.62419 27.62419 27.62419 27.62419 27.62419 27.62419 27.62420 27.62420
#> 19 17.02589 17.02573 17.02556 17.02540 17.02524 17.02509 17.02494 17.02480
#> 20 14.28605 14.28623 14.28640 14.28657 14.28673 14.28689 14.28704 14.28719
#> 21 24.60842 24.60838 24.60835 24.60832 24.60829 24.60826 24.60823 24.60820
#>      0.0065   0.0061   0.0056   0.0053
#> 1  22.10402 22.10401 22.10399 22.10397
#> 2  21.64885 21.64882 21.64879 21.64876
#> 3  26.93893 26.93895 26.93896 26.93898
#> 4  14.98088 14.98094 14.98100 14.98106
#> 5  22.60783 22.60773 22.60764 22.60754
#> 6  25.10580 25.10582 25.10584 25.10585
#> 7  19.55485 19.55478 19.55471 19.55464
#> 8  13.80120 13.80103 13.80087 13.80072
#> 9  15.83867 15.83861 15.83854 15.83848
#> 10 16.18796 16.18792 16.18787 16.18783
#> 11 12.06792 12.06799 12.06806 12.06813
#> 12 28.39863 28.39864 28.39866 28.39867
#> 13 29.48024 29.48024 29.48023 29.48023
#> 14 30.38137 30.38145 30.38152 30.38160
#> 15 18.64847 18.64854 18.64862 18.64869
#> 16 18.80843 18.80859 18.80874 18.80889
#> 17 28.90157 28.90160 28.90162 28.90165
#> 18 27.62420 27.62420 27.62420 27.62420
#> 19 17.02466 17.02452 17.02439 17.02426
#> 20 14.28733 14.28747 14.28761 14.28774
#> 21 24.60818 24.60815 24.60813 24.60811
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 22.8 14.3 24.4 22.8 17.8 16.4 17.3 15.2 14.7 32.4 30.4 33.9 15.5
#> [16] 19.2 27.3 30.4 15.8 15.0 21.4
#> 
#> $X
#>               am       carb       cyl       disp        drat       gear
#>  [1,]  0.9534626  0.7476858  0.000000 -0.4269386  0.39258159  0.2871833
#>  [2,]  0.9534626  0.7476858  0.000000 -0.4269386  0.39258159  0.2871833
#>  [3,]  0.9534626 -1.1364823 -1.080123 -0.8747505  0.29627068  0.2871833
#>  [4,] -1.0488088  0.7476858  1.080123  1.2954148 -0.93650888 -1.2205289
#>  [5,] -1.0488088 -0.5084263 -1.080123 -0.5414751 -0.01192421  0.2871833
#>  [6,] -1.0488088 -0.5084263 -1.080123 -0.5922845  0.43110595  0.2871833
#>  [7,] -1.0488088  0.7476858  0.000000 -0.3614892  0.43110595  0.2871833
#>  [8,] -1.0488088  0.1196297  1.080123  0.5703040 -1.20617941 -1.2205289
#>  [9,] -1.0488088  0.1196297  1.080123  0.5703040 -1.20617941 -1.2205289
#> [10,] -1.0488088  0.1196297  1.080123  0.5703040 -1.20617941 -1.2205289
#> [11,] -1.0488088  0.7476858  1.080123  1.9843562 -0.89798452 -1.2205289
#> [12,]  0.9534626 -1.1364823 -1.080123 -1.1270753  0.73930084  0.2871833
#> [13,]  0.9534626 -0.5084263 -1.080123 -1.1529106  2.37658619  0.2871833
#> [14,]  0.9534626 -1.1364823 -1.080123 -1.1925247  1.00897137  0.2871833
#> [15,] -1.0488088 -0.5084263  1.080123  0.9337206 -1.80330701 -1.2205289
#> [16,] -1.0488088 -0.5084263  1.080123  1.6398855 -1.18691723 -1.2205289
#> [17,]  0.9534626 -1.1364823 -1.080123 -1.1244918  0.73930084  0.2871833
#> [18,]  0.9534626 -0.5084263 -1.080123 -0.9858423  0.14217324  1.7948954
#> [19,]  0.9534626  0.7476858  1.080123  1.2179089  1.00897137  1.7948954
#> [20,]  0.9534626  3.2599099  1.080123  0.7873206 -0.30085692  1.7948954
#> [21,]  0.9534626 -0.5084263 -1.080123 -0.7627975  0.79708738  0.2871833
#>                hp       qsec         vs         wt
#>  [1,] -0.44233722 -0.7655785 -0.9534626 -0.4712004
#>  [2,] -0.44233722 -0.4634576 -0.9534626 -0.1924775
#>  [3,] -0.67020791  0.3943500  1.0488088 -0.7991096
#>  [4,]  1.36722414 -1.1000695 -0.9534626  0.5671789
#>  [5,] -1.08573682  1.1442572  1.0488088  0.1518272
#>  [6,] -0.64339960  2.7088119  1.0488088  0.1081060
#>  [7,] -0.26808316  0.5508054  1.0488088  0.4250849
#>  [8,]  0.49595385 -0.2584470 -0.9534626  1.1136943
#>  [9,]  0.49595385 -0.1505466 -0.9534626  0.7420638
#> [10,]  0.49595385  0.0652540 -0.9534626  0.7967154
#> [11,]  1.16616177 -0.2476569 -0.9534626  2.5073086
#> [12,] -1.03212018  0.8583214  1.0488088 -0.9302733
#> [13,] -1.21977840  0.3457948  1.0488088 -1.5696963
#> [14,] -1.04552434  1.0903071  1.0488088 -1.3292296
#> [15,]  0.09382911 -0.5443828 -0.9534626  0.5125274
#> [16,]  0.42893306 -0.4472725 -0.9534626  0.8677624
#> [17,] -1.03212018  0.5508054  1.0488088 -1.2199265
#> [18,] -0.40212475 -0.5281978  1.0488088 -1.6811855
#> [19,]  1.62190315 -1.8230016 -0.9534626  0.1299666
#> [20,]  2.57359838 -1.7690515 -0.9534626  0.5671789
#> [21,] -0.45574138  0.3889550  1.0488088 -0.2963154
#> attr(,"center")
#>  [1]   0.5238095   2.8095238   6.0000000 209.5761905   3.6961905   3.8095238
#>  [7] 143.0000000  17.8790476   0.4761905   3.0510952
#> attr(,"scale")
#>  [1]   0.4994328   1.5922146   1.8516402 116.1201837   0.5191520   0.6632566
#>  [7]  74.6037150   1.8535626   0.4994328   0.9148873
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 24
#> 
#> $lambda.min
#> [1] 1.05824
#> 
#> $null.dev
#> [1] 37.51297
#> 
#> $Bias
#> [1] 7.827567
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 11.04896