Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 50.75606 47.62319 40.74908 35.14328 30.61546 27.00177 24.16129 21.97290
#>   [9] 20.33267 19.15146 18.35294 17.87185 17.65246 17.64726 17.81585 18.11506
#>  [17] 18.44753 18.70375 18.76661 18.83280 18.84164 18.78111 18.43658 18.08487
#>  [25] 17.75776 17.28369 16.88699 16.57856 16.34440 16.24901 16.45290 16.67870
#>  [33] 16.39432 17.19532 16.73613 16.89052 16.38995 16.00319 14.37339 14.48463
#>  [41] 13.22678 13.22107 13.07603 13.07586 13.66998 13.52843 19.61620 20.22155
#>  [49] 20.22106 20.21192 20.83742 20.02182 41.89336 41.47393 41.47510 44.60812
#>  [57] 44.61054 44.61291 44.51281 44.51501 44.26592 44.26804 43.53306 43.20329
#>  [65] 43.20546 42.89049 42.89252 42.89454 42.89653 42.78379 42.78571 42.78760
#>  [73] 42.78947 42.79132 44.31286 44.54844 44.55295 44.19403 44.15998 44.12909
#>  [81] 44.07417 44.07616 44.07813 44.08006 44.07847 44.08178 44.35176 44.35083
#>  [89] 44.34990 44.34897 44.34805 44.25502 44.25411 44.25321 44.25231 44.25140
#>  [97] 44.25050 44.24960 44.24870 44.24781
#> 
#> $cvse
#>   [1] 12.484444 11.888733 10.304437  9.061824  8.081760  7.297550  6.656332
#>   [8]  6.119211  5.660096  5.263692  4.923181  4.637889  4.411015  4.247357
#>  [15]  4.151091  4.124861  4.178558  4.249437  4.245024  4.241097  4.235457
#>  [22]  4.228004  4.195751  4.173786  4.157833  4.127278  4.092722  4.057884
#>  [29]  4.022806  3.980157  3.902456  3.950756  3.919506  4.052888  4.171016
#>  [36]  4.043950  3.941827  3.873903  3.390627  3.373664  3.198607  3.197189
#>  [43]  3.198032  3.198014  3.373511  3.098168  7.470219  7.436629  7.435893
#>  [50]  7.443319  7.385012  7.404230 20.867709 20.890065 20.891368 22.266442
#>  [57] 22.268120 22.269778 22.275812 22.277433 22.287011 22.288593 22.358828
#>  [64] 22.374428 22.375941 22.391044 22.392526 22.393989 22.395435 22.399325
#>  [71] 22.400736 22.402130 22.403508 22.404869 23.097950 23.070939 23.073644
#>  [78] 22.901382 22.906127 22.908681 22.911634 22.912970 22.914288 22.915589
#>  [85] 22.917203 22.919429 22.858341 22.857314 22.856293 22.855279 22.854271
#>  [92] 22.857987 22.856992 22.856003 22.855021 22.854045 22.853076 22.852113
#>  [99] 22.851156 22.850206
#> 
#> $fold
#>  [1]  7  1  7  3  5  2 10  8 10  2  4  6  3  9  1  8  9  5  4  6  1
#> 
#> $lambda
#>   [1] 5.937443977 5.537280125 5.164085977 4.816043866 4.491458628 4.188749349
#>   [7] 3.906441662 3.643160568 3.397623739 3.168635271 2.955079858 2.755917366
#>  [13] 2.570177759 2.396956381 2.235409543 2.084750421 1.944245220 1.813209600
#>  [19] 1.691005341 1.577037241 1.470750209 1.371626567 1.279183527 1.192970839
#>  [25] 1.112568597 1.037585197 0.967655428 0.902438690 0.841617343 0.784895151
#>  [31] 0.731995845 0.682661775 0.636652656 0.593744398 0.553728012 0.516408598
#>  [37] 0.481604387 0.449145863 0.418874935 0.390644166 0.364316057 0.339762374
#>  [43] 0.316863527 0.295507986 0.275591737 0.257017776 0.239695638 0.223540954
#>  [49] 0.208475042 0.194424522 0.181320960 0.169100534 0.157703724 0.147075021
#>  [55] 0.137162657 0.127918353 0.119297084 0.111256860 0.103758520 0.096765543
#>  [61] 0.090243869 0.084161734 0.078489515 0.073199584 0.068266177 0.063665266
#>  [67] 0.059374440 0.055372801 0.051640860 0.048160439 0.044914586 0.041887493
#>  [73] 0.039064417 0.036431606 0.033976237 0.031686353 0.029550799 0.027559174
#>  [79] 0.025701778 0.023969564 0.022354095 0.020847504 0.019442452 0.018132096
#>  [85] 0.016910053 0.015770372 0.014707502 0.013716266 0.012791835 0.011929708
#>  [91] 0.011125686 0.010375852 0.009676554 0.009024387 0.008416173 0.007848952
#>  [97] 0.007319958 0.006826618 0.006366527 0.005937444
#> 
#> $fit
#> $beta
#>               5.9374     5.5373    5.1641    4.8160    4.4915    4.1887
#> (Intercept) 21.41905 23.0244710 24.521694 25.918010 27.220218 28.434662
#> am           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000 -0.5249177 -1.014458 -1.471004 -1.896781 -2.293862
#>                3.9064    3.6432    3.3976    3.1686    2.9551   2.7559
#> (Intercept) 29.567256 30.623518 31.608591 32.527273 33.384039 34.18306
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> wt          -2.664181 -3.009542 -3.331627 -3.632004 -3.912137 -4.17339
#>                2.5702    2.3970    2.2354    2.0848    1.9442    1.8132
#> (Intercept) 34.928234 35.623183 36.271295 36.875727 37.299395 37.299395
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -4.417035 -4.644259 -4.856169 -5.053798 -5.192322 -5.192322
#>                1.6910    1.5770      1.4708     1.3716     1.2792     1.1930
#> (Intercept) 37.299395 37.299395 36.45648718 34.7516923 33.1616209 31.6787150
#> am           0.000000  0.000000  0.00000000  0.0000000  0.0000000  0.0000000
#> carb         0.000000  0.000000  0.00000000  0.0000000  0.0000000  0.0000000
#> cyl          0.000000  0.000000  0.00000000  0.0000000  0.0000000  0.0000000
#> disp         0.000000  0.000000  0.00000000  0.0000000  0.0000000  0.0000000
#> drat         0.000000  0.000000  0.00000000  0.0000000  0.0000000  0.0000000
#> gear         0.000000  0.000000  0.00000000  0.0000000  0.0000000  0.0000000
#> hp           0.000000  0.000000  0.00000000  0.0000000  0.0000000  0.0000000
#> qsec         0.000000  0.000000  0.04464118  0.1349287  0.2191404  0.2976765
#> vs           0.000000  0.000000  0.00000000  0.0000000  0.0000000  0.0000000
#> wt          -5.192322 -5.192322 -5.17897832 -5.1519897 -5.1268173 -5.1033415
#>                 1.1126     1.0376     0.9677     0.9024     0.8416     0.7849
#> (Intercept) 30.2957520 29.0059960 27.8031653 26.6814013 25.6352405 24.6595874
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.3709195  0.4392262  0.5029293  0.5623389  0.6177446  0.6694161
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.0814478 -5.0610297 -5.0419877 -5.0242291 -5.0076673 -4.9922218
#>                 0.7320     0.6827     0.6367    0.5937    0.5537    0.5164
#> (Intercept) 23.7496901 22.9011168 20.0825939 11.511460 10.766953 10.764906
#> am           0.0000000  0.0000000  0.6154732  3.244395  3.521651  3.522290
#> carb         0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> cyl          0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> disp         0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> drat         0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> gear         0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> hp           0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> qsec         0.7176051  0.7625464  0.8619251  1.102749  1.119703  1.119759
#> vs           0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> wt          -4.9778172 -4.9643834 -4.7320636 -3.794636 -3.698296 -3.698069
#>                0.4816      0.4491     0.4189     0.3906     0.3643     0.3398
#> (Intercept) 10.763613 11.28383724 15.5036454 16.5247322 16.5244562 16.5242514
#> am           3.522693  3.59748223  4.2121359  4.3602765  4.3609280  4.3614098
#> carb         0.000000 -0.07749714 -0.7090152 -0.8616497 -0.8618398 -0.8619802
#> cyl          0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         1.119795  1.08838906  0.8330713  0.7713337  0.7713009  0.7712767
#> vs           0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -3.697925 -3.63116761 -3.0852709 -2.9534768 -2.9531458 -2.9529011
#>                 0.3169     0.2955     0.2756     0.2570     0.2397     0.2235
#> (Intercept) 16.5240995 16.5239872 16.5239040 16.5238425 16.5237970 16.5237634
#> am           4.3617659  4.3620292  4.3622238  4.3623677  4.3624740  4.3625527
#> carb        -0.8620840 -0.8621607 -0.8622174 -0.8622593 -0.8622903 -0.8623132
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.7712588  0.7712456  0.7712359  0.7712286  0.7712233  0.7712194
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.9527201 -2.9525864 -2.9524876 -2.9524145 -2.9523605 -2.9523206
#>                 0.2085     0.1944       0.1813       0.1691       0.1577
#> (Intercept) 16.5237385 16.5237202 -47.89232429 -47.89729352 -47.90180896
#> am           4.3626108  4.3626537   5.13277884   5.13348587   5.13412677
#> carb        -0.8623301 -0.8623426  -2.10729857  -2.10750316  -2.10769207
#> cyl          0.0000000  0.0000000   3.15380637   3.15391211   3.15400625
#> disp         0.0000000  0.0000000   0.00000000   0.00000000   0.00000000
#> drat         0.0000000  0.0000000   3.23723332   3.23734361   3.23744763
#> gear         0.0000000  0.0000000   3.63065079   3.63077483   3.63089136
#> hp           0.0000000  0.0000000   0.02419242   0.02420204   0.02421081
#> qsec         0.7712165  0.7712143   2.08605685   2.08619860   2.08632582
#> vs           0.0000000  0.0000000   0.00000000   0.00000000   0.00000000
#> wt          -2.9522911 -2.9522692  -3.97058664  -3.97064699  -3.97069680
#>                   0.1471     0.1372     0.1279     0.1193     0.1113     0.1038
#> (Intercept) -47.90590596 -46.414807 -46.419290 -46.423465 -46.427355 -46.430978
#> am            5.13470665   3.335184   3.335440   3.335677   3.335898   3.336103
#> carb         -2.10786629  -1.835897  -1.835956  -1.836010  -1.836061  -1.836109
#> cyl           3.15408988   3.733756   3.734085   3.734391   3.734676   3.734941
#> disp          0.00000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> drat          3.23754560   3.953985   3.954073   3.954155   3.954232   3.954303
#> gear          3.63100075   4.575386   4.575597   4.575793   4.575977   4.576148
#> hp            0.02421881   0.000000   0.000000   0.000000   0.000000   0.000000
#> qsec          2.08643974   1.434380   1.434476   1.434566   1.434649   1.434726
#> vs            0.00000000   3.218762   3.218966   3.219155   3.219332   3.219497
#> wt           -3.97073723  -3.081707  -3.081808  -3.081903  -3.081991  -3.082073
#>                 0.0968     0.0902     0.0842     0.0785     0.0732     0.0683
#> (Intercept) -46.434355 -46.437500 -46.440431 -46.443163 -46.445708 -46.448079
#> am            3.336294   3.336471   3.336636   3.336789   3.336932   3.337065
#> carb         -1.836154  -1.836195  -1.836234  -1.836269  -1.836303  -1.836334
#> cyl           3.735189   3.735419   3.735634   3.735834   3.736021   3.736194
#> disp          0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> drat          3.954371   3.954433   3.954492   3.954546   3.954597   3.954645
#> gear          4.576307   4.576456   4.576595   4.576725   4.576846   4.576958
#> hp            0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> qsec          1.434799   1.434866   1.434928   1.434986   1.435040   1.435091
#> vs            3.219652   3.219795   3.219930   3.220055   3.220172   3.220281
#> wt           -3.082149  -3.082220  -3.082286  -3.082347  -3.082404  -3.082458
#>                 0.0637     0.0594     0.0554     0.0516     0.0482     0.0449
#> (Intercept) -46.450289 -46.452349 -46.454269 -46.456059 -46.457726 -46.459281
#> am            3.337188   3.337303   3.337410   3.337509   3.337602   3.337688
#> carb         -1.836363  -1.836391  -1.836416  -1.836439  -1.836461  -1.836482
#> cyl           3.736356   3.736507   3.736648   3.736779   3.736901   3.737015
#> disp          0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> drat          3.954689   3.954731   3.954770   3.954806   3.954840   3.954871
#> gear          4.577064   4.577162   4.577253   4.577339   4.577418   4.577493
#> hp            0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> qsec          1.435138   1.435182   1.435222   1.435260   1.435295   1.435328
#> vs            3.220383   3.220478   3.220567   3.220649   3.220727   3.220799
#> wt           -3.082507  -3.082553  -3.082596  -3.082637  -3.082674  -3.082709
#>                 0.0419     0.0391     0.0364     0.0340     0.0317      0.0296
#> (Intercept) -46.460730 -46.462081 -46.463340 -46.464514 -46.465608 -43.7178812
#> am            3.337768   3.337843   3.337912   3.337977   3.338037   4.1351966
#> carb         -1.836501  -1.836519  -1.836535  -1.836551  -1.836565  -1.4887375
#> cyl           3.737121   3.737220   3.737313   3.737399   3.737479   3.0664434
#> disp          0.000000   0.000000   0.000000   0.000000   0.000000   0.0148721
#> drat          3.954900   3.954928   3.954954   3.954978   3.955000   3.5506244
#> gear          4.577562   4.577626   4.577687   4.577743   4.577795   4.0779729
#> hp            0.000000   0.000000   0.000000   0.000000   0.000000   0.0000000
#> qsec          1.435359   1.435388   1.435414   1.435439   1.435462   1.6555213
#> vs            3.220866   3.220929   3.220987   3.221042   3.221092   2.6983489
#> wt           -3.082741  -3.082771  -3.082800  -3.082826  -3.082850  -4.3223547
#>                   0.0276       0.0257       0.0240       0.0224       0.0208
#> (Intercept) -43.71408495 -43.71035793 -43.70669894 -43.70310679 -43.69958030
#> am            4.13540910   4.13561756   4.13582206   4.13602268   4.13621948
#> carb         -1.48860932  -1.48848353  -1.48836008  -1.48823893  -1.48812003
#> cyl           3.06607675   3.06571687   3.06536362   3.06501688   3.06467654
#> disp          0.01487564   0.01487912   0.01488253   0.01488587   0.01488915
#> drat          3.55035768   3.55009584   3.54983885   3.54958661   3.54933905
#> gear          4.07762723   4.07728793   4.07695489   4.07662798   4.07630710
#> hp            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> qsec          1.65554279   1.65556387   1.65558452   1.65560476   1.65562458
#> vs            2.69798819   2.69763416   2.69728665   2.69694556   2.69661075
#> wt           -4.32263344  -4.32290693  -4.32317524  -4.32343846  -4.32369668
#>                   0.0194        0.0181        0.0169       0.0158        0.0147
#> (Intercept) -43.69611832 -45.367256216 -45.363710845 -45.36025137 -45.356878689
#> am            4.13641254   4.547447396   4.547799610   4.54812341   4.548419780
#> carb         -1.48800336  -1.541048216  -1.541149849  -1.54125076  -1.541350995
#> cyl           3.06434248   3.026917564   3.026567048   3.02623267   3.025914134
#> disp          0.01489237   0.015206387   0.015203194   0.01519994   0.015196636
#> drat          3.54909607   3.471635863   3.471365281   3.47110898   3.470866625
#> gear          4.07599215   3.973529286   3.973093652   3.97267755   3.972280571
#> hp            0.00000000   0.005925376   0.005933105   0.00594050   0.005947572
#> qsec          1.65564400   1.787233860   1.787274206   1.78730813   1.787335896
#> vs            2.69628213   2.381084662   2.380219740   2.37939133   2.378598520
#> wt           -4.32395001  -4.521517120  -4.521356607  -4.52118405  -4.520999974
#>                    0.0137        0.0128        0.0119        0.0111
#> (Intercept) -45.353593517 -45.350396400 -45.347287722 -45.344267718
#> am            4.548689705   4.548934124   4.549153954   4.549350085
#> carb         -1.541450590  -1.541549581  -1.541648001  -1.541745876
#> cyl           3.025611152   3.025323426   3.025050658   3.024792549
#> disp          0.015193277   0.015189870   0.015186416   0.015182919
#> drat          3.470637901   3.470422483   3.470220051   3.470030289
#> gear          3.971902314   3.971542374   3.971200347   3.970875834
#> hp            0.005954329   0.005960783   0.005966942   0.005972816
#> qsec          1.787357771   1.787374006   1.787384844   1.787390522
#> vs            2.377840407   2.377116109   2.376424754   2.375765488
#> wt           -4.520804899  -4.520599325  -4.520383738  -4.520158611
#>                    0.0104        0.0097        0.0090        0.0084
#> (Intercept) -45.341336484 -45.338493982 -45.335740054 -45.333074426
#> am            4.549523382   4.549674684   4.549804809   4.549914547
#> carb         -1.541843234  -1.541940097  -1.542036484  -1.542132415
#> cyl           3.024548797   3.024319100   3.024103155   3.023900660
#> disp          0.015179381   0.015175806   0.015172194   0.015168550
#> drat          3.469852882   3.469687520   3.469533894   3.469391699
#> gear          3.970568436   3.970277756   3.970003402   3.969744982
#> hp            0.005978414   0.005983745   0.005988817   0.005993638
#> qsec          1.787391265   1.787387295   1.787378822   1.787366052
#> vs            2.375137468   2.374539870   2.373971880   2.373432703
#> wt           -4.519924403  -4.519681562  -4.519430519  -4.519171696
#>                    0.0078       0.0073        0.0068        0.0064
#> (Intercept) -45.330496718 -45.32800645 -45.325603054 -45.323285870
#> am            4.550004668   4.55007592   4.550129028   4.550164696
#> carb         -1.542227903  -1.54232296  -1.542417606  -1.542511841
#> cyl           3.023711314   3.02353481   3.023370859   3.023219152
#> disp          0.015164876   0.01516117   0.015157444   0.015153692
#> drat          3.469260635   3.46914040   3.469030710   3.468931265
#> gear          3.969502109   3.96927440   3.969061475   3.968862957
#> hp            0.005998216   0.00600256   0.006006676   0.006010572
#> qsec          1.787349181   1.78732840   1.787303894   1.787275840
#> vs            2.372921553   2.37243766   2.371980281   2.371548662
#> wt           -4.518905502  -4.51863233  -4.518352572  -4.518066595
#>                    0.0059
#> (Intercept) -45.321054164
#> am            4.550183608
#> carb         -1.542605676
#> cyl           3.023079394
#> disp          0.015149919
#> drat          3.468841782
#> gear          3.968678475
#> hp            0.006014256
#> qsec          1.787244409
#> vs            2.371142083
#> wt           -4.517774763
#> 
#> $iter
#>   [1]   0   5   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   4   3   3   3   3   3   3   3   3   3   3   3  73  65  11   1
#>  [37]   1  50  94  17   1   1   1   1   1   1   1   1   1   1 180   1   1   1
#>  [55] 158   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1 264   1   1   1   1   1   1  99   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.937443977 5.537280125 5.164085977 4.816043866 4.491458628 4.188749349
#>   [7] 3.906441662 3.643160568 3.397623739 3.168635271 2.955079858 2.755917366
#>  [13] 2.570177759 2.396956381 2.235409543 2.084750421 1.944245220 1.813209600
#>  [19] 1.691005341 1.577037241 1.470750209 1.371626567 1.279183527 1.192970839
#>  [25] 1.112568597 1.037585197 0.967655428 0.902438690 0.841617343 0.784895151
#>  [31] 0.731995845 0.682661775 0.636652656 0.593744398 0.553728012 0.516408598
#>  [37] 0.481604387 0.449145863 0.418874935 0.390644166 0.364316057 0.339762374
#>  [43] 0.316863527 0.295507986 0.275591737 0.257017776 0.239695638 0.223540954
#>  [49] 0.208475042 0.194424522 0.181320960 0.169100534 0.157703724 0.147075021
#>  [55] 0.137162657 0.127918353 0.119297084 0.111256860 0.103758520 0.096765543
#>  [61] 0.090243869 0.084161734 0.078489515 0.073199584 0.068266177 0.063665266
#>  [67] 0.059374440 0.055372801 0.051640860 0.048160439 0.044914586 0.041887493
#>  [73] 0.039064417 0.036431606 0.033976237 0.031686353 0.029550799 0.027559174
#>  [79] 0.025701778 0.023969564 0.022354095 0.020847504 0.019442452 0.018132096
#>  [85] 0.016910053 0.015770372 0.014707502 0.013716266 0.012791835 0.011929708
#>  [91] 0.011125686 0.010375852 0.009676554 0.009024387 0.008416173 0.007848952
#>  [97] 0.007319958 0.006826618 0.006366527 0.005937444
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 32
#> 
#> $loss
#>   [1] 944.91238 802.79370 683.89029 584.86129 502.82237 435.28480 380.10260
#>   [8] 335.42663 299.66479 271.44740 249.59725 233.10358 221.09951 212.84246
#>  [15] 207.69723 205.12125 204.59432 204.59432 204.59432 204.59432 199.73927
#>  [22] 190.67071 183.11815 176.86240 171.71347 167.50750 164.10332 161.37938
#>  [29] 159.23111 157.56865 156.31484 155.40350 146.51341 129.09101 128.93284
#>  [36] 128.93284 128.93284 126.91288 117.53734 117.16518 117.16518 117.16517
#>  [43] 117.16517 117.16517 117.16517 117.16517 117.16517 117.16517 117.16517
#>  [50] 117.16517  80.52374  80.52372  80.52371  80.52370  76.57609  76.57608
#>  [57]  76.57607  76.57606  76.57606  76.57605  76.57605  76.57604  76.57604
#>  [64]  76.57604  76.57604  76.57603  76.57603  76.57603  76.57603  76.57603
#>  [71]  76.57603  76.57602  76.57602  76.57602  76.57602  76.57602  75.21659
#>  [78]  75.21657  75.21655  75.21653  75.21651  75.21649  75.21648  75.07288
#>  [85]  75.07286  75.07283  75.07281  75.07279  75.07277  75.07275  75.07273
#>  [92]  75.07272  75.07270  75.07269  75.07267  75.07266  75.07264  75.07263
#>  [99]  75.07262  75.07261
#> 
#> $linear.predictors
#>      5.9374   5.5373   5.1641   4.8160   4.4915   4.1887   3.9064   3.6432
#> 1  21.41905 21.64919 21.86381 22.06398 22.25065 22.42474 22.58710 22.73852
#> 2  21.41905 21.51533 21.60513 21.68887 21.76697 21.83981 21.90773 21.97108
#> 3  21.41905 21.80666 22.16815 22.50528 22.81969 23.11290 23.38636 23.64138
#> 4  21.41905 21.33686 21.26021 21.18873 21.12207 21.05989 21.00191 20.94784
#> 5  21.41905 21.15051 20.90008 20.66652 20.44871 20.24557 20.05613 19.87945
#> 6  21.41905 21.37098 21.32615 21.28435 21.24536 21.20900 21.17508 21.14346
#> 7  21.41905 20.26865 19.19579 18.19524 17.26212 16.39188 15.58030 14.82342
#> 8  21.41905 20.17732 19.01928 17.93928 16.93208 15.99275 15.11674 14.29976
#> 9  21.41905 20.21879 19.09942 18.05549 17.08192 16.17397 15.32721 14.53751
#> 10 21.41905 21.86965 22.28989 22.68180 23.04730 23.38816 23.70606 24.00252
#> 11 21.41905 22.17673 22.88334 23.54234 24.15692 24.73007 25.26460 25.76311
#> 12 21.41905 22.06125 22.66016 23.21872 23.73962 24.22542 24.67848 25.10101
#> 13 21.41905 21.73055 22.02106 22.29198 22.54465 22.78029 23.00005 23.20500
#> 14 21.41905 21.17676 20.95080 20.74007 20.54355 20.36027 20.18934 20.02993
#> 15 21.41905 21.00879 20.62618 20.26935 19.93658 19.62623 19.33680 19.06688
#> 16 21.41905 21.00616 20.62110 20.26200 19.92709 19.61476 19.32348 19.05183
#> 17 21.41905 22.00876 22.55872 23.07162 23.54995 23.99604 24.41207 24.80005
#> 18 21.41905 21.90115 22.35075 22.77006 23.16111 23.52580 23.86591 24.18310
#> 19 21.41905 22.23027 22.98682 23.69238 24.35039 24.96405 25.53635 26.07008
#> 20 21.41905 21.57045 21.71165 21.84333 21.96613 22.08066 22.18747 22.28709
#> 21 21.41905 21.56520 21.70150 21.82862 21.94717 22.05772 22.16083 22.25699
#>      3.3976   3.1686   2.9551   2.7559   2.5702   2.3970    2.2354    2.0848
#> 1  22.87973 23.01142 23.13424 23.24878 23.35560 23.45522 23.548131 23.634777
#> 2  22.03016 22.08526 22.13665 22.18457 22.22926 22.27094 22.309808 22.346059
#> 3  23.87922 24.10102 24.30788 24.50080 24.68071 24.84850 25.004982 25.150916
#> 4  20.89741 20.85038 20.80652 20.76561 20.72747 20.69189 20.658710 20.627767
#> 5  19.71468 19.56102 19.41771 19.28406 19.15942 19.04318 18.934770 18.833669
#> 6  21.11397 21.08646 21.06081 21.03688 21.01457 20.99377 20.974361 20.956264
#> 7  14.11755 13.45925 12.84532 12.27277 11.73880 11.24082 10.776405 10.343289
#> 8  13.53785 12.82728 12.16461 11.54660 10.97024 10.43272  9.931432  9.463928
#> 9  13.80104 13.11421 12.47367 11.87629 11.31918 10.79962 10.315069  9.863178
#> 10 24.27901 24.53686 24.77734 25.00160 25.21076 25.40581 25.587722 25.757372
#> 11 26.22801 26.66159 27.06594 27.44304 27.79472 28.12270 28.428581 28.713844
#> 12 25.49506 25.86255 26.20527 26.52489 26.82297 27.10097 27.360224 27.602008
#> 13 23.39613 23.57438 23.74062 23.89566 24.04024 24.17508 24.300837 24.418116
#> 14 19.88126 19.74262 19.61332 19.49273 19.38027 19.27539 19.177579 19.086359
#> 15 18.81514 18.58038 18.36143 18.15725 17.96682 17.78923 17.623604 17.469144
#> 16 18.79849 18.56222 18.34187 18.13638 17.94473 17.76601 17.599323 17.443875
#> 17 25.16189 25.49934 25.81405 26.10755 26.38127 26.63654 26.874607 27.096628
#> 18 24.47891 24.75478 25.01207 25.25201 25.47578 25.68447 25.879092 26.060600
#> 19 26.56784 27.03205 27.46498 27.86872 28.24526 28.59642 28.923911 29.229331
#> 20 22.37998 22.46662 22.54742 22.62277 22.69305 22.75858 22.819706 22.876707
#> 21 22.34667 22.43030 22.50830 22.58104 22.64888 22.71214 22.771144 22.826169
#>       1.9442    1.8132    1.6910    1.5770    1.4708    1.3716    1.2792
#> 1  23.695510 23.695510 23.695510 23.695510 23.622358 23.474406 23.336411
#> 2  22.371468 22.371468 22.371468 22.371468 22.326717 22.236209 22.151791
#> 3  25.253207 25.253207 25.253207 25.253207 25.272030 25.310100 25.345608
#> 4  20.606078 20.606078 20.606078 20.606078 20.673896 20.811060 20.938993
#> 5  18.762804 18.762804 18.762804 18.762804 18.674651 18.496360 18.330067
#> 6  20.943579 20.943579 20.943579 20.943579 21.164988 21.612792 22.030462
#> 7  10.039702 10.039702 10.039702 10.039702 10.069499 10.129765 10.185974
#> 8   9.136238  9.136238  9.136238  9.136238  9.161215  9.211730  9.258846
#> 9   9.546432  9.546432  9.546432  9.546432  9.552497  9.564766  9.576208
#> 10 25.876285 25.876285 25.876285 25.876285 25.931899 26.044377 26.149286
#> 11 28.913794 28.913794 28.913794 28.913794 28.919192 28.930109 28.940291
#> 12 27.771483 27.771483 27.771483 27.771483 27.841421 27.982873 28.114805
#> 13 24.500320 24.500320 24.500320 24.500320 24.583576 24.751961 24.909016
#> 14 19.022420 19.022420 19.022420 19.022420 18.979580 18.892936 18.812123
#> 15 17.360877 17.360877 17.360877 17.360877 17.257131 17.047303 16.851596
#> 16 17.334915 17.334915 17.334915 17.334915 17.304448 17.242827 17.185352
#> 17 27.252251 27.252251 27.252251 27.252251 27.278882 27.332745 27.382983
#> 18 26.187825 26.187825 26.187825 26.187825 26.118981 25.979744 25.849877
#> 19 29.443411 29.443411 29.443411 29.443411 29.375129 29.237027 29.108219
#> 20 22.916662 22.916662 22.916662 22.916662 22.802656 22.572076 22.357013
#> 21 22.864739 22.864739 22.864739 22.864739 22.889253 22.938835 22.985080
#>       1.1930    1.1126    1.0376    0.9677    0.9024    0.8416    0.7849
#> 1  23.207716 23.087694 22.975762 22.871373 22.774020 22.683228 22.598556
#> 2  22.073062 21.999640 21.931166 21.867307 21.807752 21.752210 21.700412
#> 3  25.378722 25.409605 25.438407 25.465268 25.490318 25.513680 25.535467
#> 4  21.058303 21.169573 21.273343 21.370120 21.460374 21.544545 21.623044
#> 5  18.174982 18.030349 17.895463 17.769669 17.652353 17.542943 17.440907
#> 6  22.419981 22.783249 23.122033 23.437984 23.732642 24.007440 24.263718
#> 7  10.238396 10.287284 10.332878 10.375398 10.415053 10.452035 10.486525
#> 8   9.302786  9.343765  9.381982  9.417624  9.450863  9.481862  9.510772
#> 9   9.586880  9.596832  9.606113  9.614769  9.622841  9.630370  9.637391
#> 10 26.247125 26.338370 26.423465 26.502825 26.576837 26.645860 26.710231
#> 11 28.949787 28.958643 28.966903 28.974605 28.981789 28.988488 28.994736
#> 12 28.237846 28.352594 28.459608 28.559410 28.652486 28.739289 28.820241
#> 13 25.055485 25.192083 25.319475 25.438280 25.549079 25.652410 25.748777
#> 14 18.736756 18.666468 18.600918 18.539785 18.482773 18.429603 18.380017
#> 15 16.669079 16.498862 16.340118 16.192073 16.054005 15.925243 15.805158
#> 16 17.131751 17.081763 17.035144 16.991667 16.951120 16.913305 16.878040
#> 17 27.429835 27.473530 27.514279 27.552282 27.587724 27.620777 27.651603
#> 18 25.728762 25.615810 25.510470 25.412230 25.320612 25.235168 25.155482
#> 19 28.988092 28.876061 28.771581 28.674143 28.583271 28.498524 28.419488
#> 20 22.156445 21.969394 21.794950 21.632263 21.480541 21.339044 21.207083
#> 21 23.028209 23.068430 23.105941 23.140924 23.173549 23.203975 23.232351
#>       0.7320    0.6827    0.6367   0.5937   0.5537   0.5164   0.4816   0.4491
#> 1  22.519590 22.445946 22.487347 22.96515 23.02938 23.02950 23.02957 22.97256
#> 2  21.652105 21.607054 21.763349 22.61506 22.71335 22.71355 22.71369 22.65611
#> 3  25.555786 25.574735 25.760105 26.47445 26.54623 26.54640 26.54651 26.63443
#> 4  21.696252 21.764526 21.624833 20.74914 20.64395 20.64374 20.64360 20.69042
#> 5  17.345748 17.257003 16.842020 15.43215 15.30013 15.29979 15.29958 15.25066
#> 6  24.502724 24.725621 24.914678 24.81130 24.75852 24.75848 24.75846 24.61477
#> 7  10.518690 10.548688 10.736673 11.41704 11.48316 11.48332 11.48342 11.47945
#> 8   9.537733  9.562878  9.775386 10.58033 10.66050 10.66070 10.66082 10.67349
#> 9   9.643939  9.650045  9.804449 10.43901 10.50479 10.50494 10.50504 10.52500
#> 10 26.770264 26.826251 27.069209 27.87817 27.95297 27.95316 27.95328 28.00619
#> 11 29.000563 29.005997 29.018637 29.05042 29.05275 29.05276 29.05276 29.01895
#> 12 28.895738 28.966146 29.167040 29.73739 29.78432 29.78445 29.78454 29.79957
#> 13 25.838650 25.922465 25.665178 24.22368 24.05591 24.05555 24.05533 24.03418
#> 14 18.333772 18.290645 17.966406 16.75771 16.63834 16.63805 16.63786 16.70826
#> 15 15.693167 15.588724 15.193735 13.93341 13.82012 13.81982 13.81963 13.80224
#> 16 16.845151 16.814478 16.583632 15.72295 15.63794 15.63773 15.63760 15.72404
#> 17 27.680351 27.707162 27.831908 28.25518 28.29479 28.29489 28.29495 28.34807
#> 18 25.081167 25.011861 24.965600 25.05124 25.07329 25.07331 25.07333 25.13172
#> 19 28.345780 28.277039 28.104989 27.65102 27.61606 27.61595 27.61588 27.62614
#> 20 21.084016 20.969244 20.950090 21.33732 21.39972 21.39982 21.39988 21.22803
#> 21 23.258814 23.283494 23.574737 24.71789 24.83381 24.83409 24.83427 24.87572
#>      0.4189   0.3906   0.3643   0.3398   0.3169   0.2955   0.2756   0.2570
#> 1  22.50866 22.39645 22.39640 22.39635 22.39632 22.39630 22.39628 22.39627
#> 2  22.18844 22.07526 22.07527 22.07528 22.07528 22.07529 22.07529 22.07529
#> 3  27.35239 27.52581 27.52616 27.52641 27.52660 27.52674 27.52684 27.52691
#> 4  21.07039 21.16238 21.16234 21.16231 21.16229 21.16228 21.16226 21.16226
#> 5  14.84902 14.75215 14.75177 14.75150 14.75129 14.75114 14.75103 14.75095
#> 6  23.44434 23.16152 23.16116 23.16089 23.16069 23.16054 23.16043 23.16035
#> 7  11.44853 11.44096 11.44107 11.44115 11.44122 11.44126 11.44130 11.44132
#> 8  10.77841 10.80364 10.80382 10.80395 10.80404 10.80411 10.80416 10.80420
#> 9  10.68891 10.72843 10.72859 10.72871 10.72880 10.72887 10.72892 10.72895
#> 10 28.43907 28.54358 28.54385 28.54406 28.54421 28.54432 28.54440 28.54446
#> 11 28.74352 28.67694 28.67687 28.67681 28.67677 28.67674 28.67671 28.67670
#> 12 29.92341 29.95327 29.95341 29.95351 29.95359 29.95365 29.95369 29.95372
#> 13 23.85919 23.81715 23.81684 23.81662 23.81645 23.81633 23.81623 23.81617
#> 14 17.27937 17.41759 17.41755 17.41752 17.41749 17.41748 17.41746 17.41745
#> 15 13.65777 13.62303 13.62276 13.62256 13.62242 13.62231 13.62223 13.62217
#> 16 16.42661 16.59655 16.59661 16.59665 16.59669 16.59671 16.59673 16.59674
#> 17 28.78181 28.88659 28.88679 28.88695 28.88706 28.88714 28.88720 28.88725
#> 18 25.60756 25.72254 25.72270 25.72281 25.72290 25.72296 25.72301 25.72304
#> 19 27.70864 27.72864 27.72858 27.72854 27.72851 27.72848 27.72847 27.72845
#> 20 19.82809 19.48965 19.48930 19.48903 19.48884 19.48869 19.48859 19.48851
#> 21 25.21582 25.29785 25.29816 25.29838 25.29855 25.29867 25.29876 25.29883
#>      0.2397   0.2235   0.2085   0.1944   0.1813   0.1691   0.1577   0.1471
#> 1  22.39626 22.39625 22.39625 22.39625 21.47664 21.47635 21.47609 21.47584
#> 2  22.07529 22.07530 22.07530 22.07530 21.63233 21.63211 21.63190 21.63171
#> 3  27.52697 27.52701 27.52704 27.52706 26.59398 26.59425 26.59450 26.59473
#> 4  21.16225 21.16224 21.16224 21.16224 20.23452 20.23431 20.23412 20.23395
#> 5  14.75089 14.75084 14.75081 14.75078 14.98769 14.98786 14.98802 14.98815
#> 6  23.16029 23.16025 23.16022 23.16019 25.28250 25.28244 25.28239 25.28234
#> 7  11.44134 11.44135 11.44136 11.44137 10.90715 10.90710 10.90707 10.90703
#> 8  10.80423 10.80425 10.80427 10.80428 10.35103 10.35105 10.35108 10.35111
#> 9  10.72898 10.72900 10.72901 10.72902 10.93773 10.93788 10.93801 10.93814
#> 10 28.54451 28.54454 28.54456 28.54458 28.95583 28.95600 28.95615 28.95628
#> 11 28.67668 28.67667 28.67667 28.67666 29.60253 29.60235 29.60218 29.60203
#> 12 29.95374 29.95376 29.95377 29.95378 31.73112 31.73137 31.73160 31.73181
#> 13 23.81612 23.81608 23.81605 23.81603 19.78648 19.78613 19.78582 19.78552
#> 14 17.41744 17.41744 17.41744 17.41743 17.79442 17.79419 17.79397 17.79377
#> 15 13.62212 13.62209 13.62207 13.62205 14.70199 14.70214 14.70228 14.70241
#> 16 16.59675 16.59675 16.59676 16.59676 18.52020 18.52024 18.52029 18.52033
#> 17 28.88728 28.88731 28.88733 28.88734 28.81899 28.81908 28.81918 28.81926
#> 18 25.72307 25.72309 25.72310 25.72311 26.67688 26.67686 26.67684 26.67684
#> 19 27.72844 27.72844 27.72843 27.72843 27.97931 27.97949 27.97966 27.97981
#> 20 19.48845 19.48841 19.48838 19.48835 18.96057 18.96045 18.96033 18.96023
#> 21 25.29888 25.29892 25.29895 25.29897 23.86811 23.86833 23.86854 23.86872
#>      0.1372   0.1279   0.1193   0.1113   0.1038   0.0968   0.0902   0.0842
#> 1  21.23723 21.23725 21.23726 21.23727 21.23728 21.23729 21.23729 21.23730
#> 2  21.25465 21.25469 21.25473 21.25476 21.25480 21.25483 21.25485 21.25488
#> 3  26.30690 26.30687 26.30684 26.30681 26.30679 26.30676 26.30674 26.30672
#> 4  21.25168 21.25176 21.25184 21.25191 21.25197 21.25203 21.25209 21.25214
#> 5  14.24899 14.24897 14.24896 14.24895 14.24894 14.24893 14.24892 14.24891
#> 6  25.00827 25.00826 25.00825 25.00824 25.00823 25.00822 25.00822 25.00821
#> 7  11.03418 11.03418 11.03417 11.03417 11.03417 11.03417 11.03416 11.03416
#> 8  10.54524 10.54521 10.54518 10.54516 10.54513 10.54511 10.54509 10.54507
#> 9  11.12436 11.12432 11.12428 11.12425 11.12421 11.12419 11.12416 11.12413
#> 10 28.81969 28.81977 28.81985 28.81992 28.81999 28.82005 28.82011 28.82017
#> 11 30.78482 30.78489 30.78495 30.78501 30.78506 30.78511 30.78516 30.78521
#> 12 31.11486 31.11503 31.11519 31.11534 31.11548 31.11561 31.11574 31.11585
#> 13 19.36452 19.36413 19.36376 19.36342 19.36311 19.36281 19.36254 19.36228
#> 14 17.77299 17.77315 17.77331 17.77346 17.77359 17.77372 17.77384 17.77395
#> 15 14.85622 14.85618 14.85615 14.85611 14.85609 14.85606 14.85603 14.85601
#> 16 18.29489 18.29508 18.29524 18.29540 18.29555 18.29568 18.29581 18.29593
#> 17 28.81875 28.81880 28.81885 28.81890 28.81895 28.81899 28.81903 28.81906
#> 18 25.93598 25.93578 25.93560 25.93543 25.93527 25.93512 25.93499 25.93486
#> 19 28.76422 28.76425 28.76428 28.76431 28.76433 28.76436 28.76438 28.76440
#> 20 19.19445 19.19442 19.19440 19.19437 19.19435 19.19433 19.19432 19.19430
#> 21 24.06711 24.06700 24.06689 24.06679 24.06670 24.06661 24.06653 24.06645
#>      0.0785   0.0732   0.0683   0.0637   0.0594   0.0554   0.0516   0.0482
#> 1  21.23731 21.23731 21.23732 21.23733 21.23733 21.23734 21.23734 21.23734
#> 2  21.25490 21.25492 21.25494 21.25496 21.25498 21.25500 21.25501 21.25503
#> 3  26.30670 26.30668 26.30667 26.30665 26.30664 26.30662 26.30661 26.30660
#> 4  21.25219 21.25224 21.25228 21.25232 21.25236 21.25239 21.25242 21.25245
#> 5  14.24890 14.24890 14.24889 14.24888 14.24888 14.24887 14.24887 14.24886
#> 6  25.00820 25.00820 25.00819 25.00819 25.00818 25.00818 25.00817 25.00817
#> 7  11.03416 11.03416 11.03416 11.03415 11.03415 11.03415 11.03415 11.03415
#> 8  10.54505 10.54503 10.54502 10.54500 10.54499 10.54498 10.54497 10.54495
#> 9  11.12411 11.12409 11.12406 11.12405 11.12403 11.12401 11.12399 11.12398
#> 10 28.82022 28.82026 28.82031 28.82035 28.82039 28.82042 28.82045 28.82048
#> 11 30.78525 30.78529 30.78532 30.78536 30.78539 30.78542 30.78544 30.78547
#> 12 31.11595 31.11605 31.11614 31.11623 31.11631 31.11638 31.11645 31.11652
#> 13 19.36205 19.36182 19.36162 19.36142 19.36124 19.36108 19.36092 19.36078
#> 14 17.77405 17.77415 17.77424 17.77432 17.77439 17.77447 17.77453 17.77460
#> 15 14.85599 14.85597 14.85595 14.85594 14.85592 14.85591 14.85589 14.85588
#> 16 18.29604 18.29614 18.29624 18.29633 18.29641 18.29649 18.29656 18.29663
#> 17 28.81910 28.81913 28.81916 28.81918 28.81921 28.81923 28.81925 28.81927
#> 18 25.93474 25.93463 25.93452 25.93443 25.93434 25.93425 25.93417 25.93410
#> 19 28.76442 28.76444 28.76446 28.76447 28.76449 28.76450 28.76452 28.76453
#> 20 19.19428 19.19427 19.19425 19.19424 19.19423 19.19422 19.19421 19.19420
#> 21 24.06638 24.06632 24.06626 24.06620 24.06615 24.06610 24.06605 24.06601
#>      0.0449   0.0419   0.0391   0.0364   0.0340   0.0317   0.0296   0.0276
#> 1  21.23735 21.23735 21.23735 21.23735 21.23736 21.23736 21.32520 21.32529
#> 2  21.25504 21.25505 21.25506 21.25507 21.25508 21.25509 21.15009 21.15012
#> 3  26.30659 26.30657 26.30656 26.30656 26.30655 26.30654 26.26207 26.26211
#> 4  21.25248 21.25251 21.25253 21.25256 21.25258 21.25260 21.18420 21.18414
#> 5  14.24886 14.24885 14.24885 14.24885 14.24884 14.24884 14.63675 14.63685
#> 6  25.00817 25.00816 25.00816 25.00816 25.00816 25.00815 24.88912 24.88903
#> 7  11.03415 11.03415 11.03414 11.03414 11.03414 11.03414 11.58951 11.58966
#> 8  10.54494 10.54493 10.54492 10.54492 10.54491 10.54490 10.64261 10.64265
#> 9  11.12397 11.12395 11.12394 11.12393 11.12392 11.12391 10.84107 10.84099
#> 10 28.82051 28.82054 28.82056 28.82059 28.82061 28.82063 28.58539 28.58532
#> 11 30.78549 30.78552 30.78554 30.78556 30.78557 30.78559 31.02590 31.02586
#> 12 31.11657 31.11663 31.11668 31.11673 31.11678 31.11682 31.25899 31.25896
#> 13 19.36064 19.36051 19.36040 19.36029 19.36019 19.36009 19.38725 19.38749
#> 14 17.77465 17.77471 17.77476 17.77481 17.77485 17.77489 17.31312 17.31297
#> 15 14.85587 14.85586 14.85585 14.85584 14.85583 14.85582 14.45544 14.45528
#> 16 18.29669 18.29675 18.29680 18.29686 18.29690 18.29695 18.56206 18.56203
#> 17 28.81929 28.81931 28.81933 28.81934 28.81936 28.81937 28.79163 28.79162
#> 18 25.93403 25.93397 25.93391 25.93386 25.93380 25.93376 26.01122 26.01130
#> 19 28.76454 28.76455 28.76456 28.76457 28.76458 28.76459 29.03261 29.03259
#> 20 19.19419 19.19418 19.19417 19.19417 19.19416 19.19415 18.97079 18.97075
#> 21 24.06597 24.06593 24.06590 24.06587 24.06584 24.06581 23.88499 23.88501
#>      0.0257   0.0240   0.0224   0.0208   0.0194   0.0181   0.0169   0.0158
#> 1  21.32538 21.32546 21.32555 21.32563 21.32571 21.26731 21.26732 21.26734
#> 2  21.15015 21.15018 21.15021 21.15024 21.15026 21.11517 21.11525 21.11533
#> 3  26.26214 26.26218 26.26221 26.26225 26.26228 26.35167 26.35191 26.35214
#> 4  21.18408 21.18402 21.18397 21.18391 21.18386 21.02970 21.02936 21.02904
#> 5  14.63695 14.63705 14.63715 14.63724 14.63734 14.84242 14.84253 14.84263
#> 6  24.88894 24.88885 24.88877 24.88869 24.88860 24.93118 24.93116 24.93115
#> 7  11.58981 11.58995 11.59010 11.59024 11.59038 11.56499 11.56487 11.56475
#> 8  10.64269 10.64273 10.64276 10.64280 10.64284 10.61208 10.61208 10.61208
#> 9  10.84091 10.84084 10.84076 10.84069 10.84061 10.83761 10.83770 10.83779
#> 10 28.58524 28.58517 28.58509 28.58502 28.58495 28.62421 28.62429 28.62437
#> 11 31.02582 31.02577 31.02573 31.02569 31.02565 30.85270 30.85221 30.85175
#> 12 31.25893 31.25890 31.25887 31.25884 31.25881 31.40761 31.40763 31.40764
#> 13 19.38773 19.38796 19.38819 19.38842 19.38864 19.36415 19.36459 19.36501
#> 14 17.31283 17.31269 17.31255 17.31241 17.31228 17.22762 17.22750 17.22738
#> 15 14.45513 14.45498 14.45484 14.45469 14.45455 14.50628 14.50631 14.50635
#> 16 18.56200 18.56198 18.56195 18.56192 18.56190 18.58581 18.58559 18.58537
#> 17 28.79160 28.79159 28.79158 28.79156 28.79155 28.80825 28.80827 28.80828
#> 18 26.01138 26.01145 26.01153 26.01160 26.01167 25.99205 25.99231 25.99255
#> 19 29.03257 29.03255 29.03254 29.03252 29.03250 29.02146 29.02118 29.02091
#> 20 18.97071 18.97067 18.97063 18.97059 18.97055 18.94976 18.94975 18.94974
#> 21 23.88502 23.88503 23.88504 23.88506 23.88507 23.90796 23.90819 23.90841
#>      0.0147   0.0137   0.0128   0.0119   0.0111   0.0104   0.0097   0.0090
#> 1  21.26735 21.26736 21.26737 21.26738 21.26738 21.26739 21.26740 21.26740
#> 2  21.11540 21.11547 21.11554 21.11561 21.11568 21.11575 21.11582 21.11588
#> 3  26.35237 26.35258 26.35279 26.35299 26.35319 26.35338 26.35356 26.35373
#> 4  21.02872 21.02842 21.02814 21.02786 21.02759 21.02734 21.02709 21.02686
#> 5  14.84273 14.84281 14.84289 14.84295 14.84300 14.84305 14.84309 14.84312
#> 6  24.93114 24.93113 24.93112 24.93112 24.93111 24.93111 24.93111 24.93111
#> 7  11.56462 11.56450 11.56437 11.56425 11.56412 11.56399 11.56386 11.56373
#> 8  10.61207 10.61207 10.61207 10.61207 10.61206 10.61206 10.61206 10.61206
#> 9  10.83788 10.83797 10.83806 10.83815 10.83824 10.83833 10.83843 10.83852
#> 10 28.62445 28.62452 28.62460 28.62467 28.62475 28.62482 28.62489 28.62496
#> 11 30.85129 30.85086 30.85043 30.85002 30.84963 30.84925 30.84888 30.84852
#> 12 31.40765 31.40765 31.40764 31.40764 31.40762 31.40761 31.40758 31.40756
#> 13 19.36543 19.36584 19.36623 19.36661 19.36699 19.36735 19.36770 19.36804
#> 14 17.22727 17.22717 17.22707 17.22698 17.22690 17.22683 17.22676 17.22670
#> 15 14.50639 14.50643 14.50647 14.50652 14.50657 14.50662 14.50668 14.50673
#> 16 18.58515 18.58494 18.58473 18.58452 18.58431 18.58411 18.58390 18.58371
#> 17 28.80829 28.80830 28.80831 28.80832 28.80833 28.80834 28.80834 28.80835
#> 18 25.99279 25.99302 25.99325 25.99347 25.99369 25.99390 25.99410 25.99430
#> 19 29.02064 29.02038 29.02013 29.01988 29.01964 29.01941 29.01918 29.01895
#> 20 18.94973 18.94973 18.94973 18.94973 18.94973 18.94974 18.94974 18.94975
#> 21 23.90862 23.90884 23.90904 23.90925 23.90945 23.90964 23.90983 23.91002
#>      0.0084   0.0078   0.0073   0.0068   0.0064   0.0059
#> 1  21.26741 21.26741 21.26741 21.26741 21.26741 21.26741
#> 2  21.11594 21.11600 21.11606 21.11612 21.11618 21.11623
#> 3  26.35390 26.35406 26.35421 26.35436 26.35450 26.35464
#> 4  21.02664 21.02642 21.02622 21.02602 21.02584 21.02566
#> 5  14.84314 14.84315 14.84316 14.84316 14.84315 14.84313
#> 6  24.93111 24.93111 24.93112 24.93112 24.93113 24.93114
#> 7  11.56360 11.56346 11.56333 11.56319 11.56306 11.56292
#> 8  10.61205 10.61205 10.61204 10.61204 10.61204 10.61203
#> 9  10.83861 10.83871 10.83880 10.83890 10.83899 10.83909
#> 10 28.62503 28.62509 28.62516 28.62523 28.62529 28.62536
#> 11 30.84818 30.84785 30.84752 30.84722 30.84692 30.84663
#> 12 31.40753 31.40750 31.40746 31.40743 31.40739 31.40734
#> 13 19.36837 19.36870 19.36901 19.36931 19.36960 19.36988
#> 14 17.22665 17.22660 17.22656 17.22652 17.22649 17.22647
#> 15 14.50679 14.50685 14.50692 14.50698 14.50705 14.50712
#> 16 18.58351 18.58332 18.58313 18.58294 18.58275 18.58257
#> 17 28.80836 28.80836 28.80837 28.80837 28.80838 28.80838
#> 18 25.99449 25.99468 25.99486 25.99504 25.99521 25.99537
#> 19 29.01873 29.01852 29.01831 29.01810 29.01790 29.01771
#> 20 18.94977 18.94978 18.94980 18.94982 18.94984 18.94986
#> 21 23.91020 23.91038 23.91056 23.91073 23.91089 23.91106
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 22.8 21.4 14.3 22.8 10.4 10.4 14.7 32.4 30.4 33.9 21.5 15.5 13.3
#> [16] 19.2 27.3 26.0 30.4 19.7 21.4
#> 
#> $X
#>               am       carb        cyl       disp        drat       gear
#>  [1,]  0.9534626  0.9966832  0.1608169 -0.4002298  0.38512503  0.4082483
#>  [2,]  0.9534626  0.9966832  0.1608169 -0.4002298  0.38512503  0.4082483
#>  [3,]  0.9534626 -1.0963516 -0.9649013 -0.7728858  0.29258927  0.4082483
#>  [4,] -1.0488088 -1.0963516  0.1608169  0.3020834 -1.13246148 -1.0206207
#>  [5,] -1.0488088  0.9966832  1.2865350  1.0330625 -0.89186849 -1.0206207
#>  [6,] -1.0488088 -0.3986733 -0.9649013 -0.5378259  0.42213934  0.4082483
#>  [7,] -1.0488088  0.9966832  1.2865350  1.8357062 -1.41006876 -1.0206207
#>  [8,] -1.0488088  0.9966832  1.2865350  1.7497087 -1.28051870 -1.0206207
#>  [9,] -1.0488088  0.9966832  1.2865350  1.6063794 -0.85485419 -1.0206207
#> [10,]  0.9534626 -1.0963516 -0.9649013 -0.9828632  0.71825378  0.4082483
#> [11,]  0.9534626 -0.3986733 -0.9649013 -1.0043625  2.29136174  0.4082483
#> [12,]  0.9534626 -1.0963516 -0.9649013 -1.0373283  0.97735391  0.4082483
#> [13,] -1.0488088 -1.0963516 -0.9649013 -0.6861716  0.01498198 -1.0206207
#> [14,] -1.0488088 -0.3986733  1.2865350  0.7320711 -1.72469036 -1.0206207
#> [15,] -1.0488088  0.9966832  1.2865350  0.9613979  0.07050344 -1.0206207
#> [16,] -1.0488088 -0.3986733  1.2865350  1.3197210 -1.13246148 -1.0206207
#> [17,]  0.9534626 -1.0963516 -0.9649013 -0.9807132  0.71825378  0.4082483
#> [18,]  0.9534626 -0.3986733 -0.9649013 -0.6847383  1.36600412  1.8371173
#> [19,]  0.9534626 -0.3986733 -0.9649013 -0.8653332  0.14453205  1.8371173
#> [20,]  0.9534626  2.3920398  0.1608169 -0.5077267 -0.13307524  1.8371173
#> [21,]  0.9534626 -0.3986733 -0.9649013 -0.6797218  0.77377523  0.4082483
#>               hp         qsec         vs          wt
#>  [1,] -0.3935657 -0.862994891 -0.9534626 -0.38340783
#>  [2,] -0.3935657 -0.542438356 -0.9534626 -0.16040914
#>  [3,] -0.6706918  0.367713237  1.0488088 -0.64575922
#>  [4,] -0.3935657  0.842823817  1.0488088  0.13692244
#>  [5,]  1.8071419 -1.217896770 -0.9534626  0.44737159
#>  [6,] -0.6380888  2.823405270  1.0488088  0.08007964
#>  [7,]  1.1550804  0.007087134 -0.9534626  1.91653941
#>  [8,]  1.3180958 -0.084500447 -0.9534626  2.06870322
#>  [9,]  1.5626188 -0.313469401 -0.9534626  1.99961735
#> [10,] -1.1108334  0.859996488  1.0488088 -0.75069978
#> [11,] -1.3390549  0.316195222  1.0488088 -1.26228500
#> [12,] -1.1271349  1.106138114  1.0488088 -1.06989398
#> [13,] -0.6054857  1.169104576  1.0488088 -0.51895605
#> [14,]  0.2584958 -0.628301713 -0.9534626  0.40364636
#> [15,]  1.8071419 -1.464038396 -0.9534626  0.68348785
#> [16,]  0.6660343 -0.525265684 -0.9534626  0.68786037
#> [17,] -1.1108334  0.533715729  1.0488088 -0.98244351
#> [18,] -0.7032949 -0.725613519 -0.9534626 -0.80317006
#> [19,] -0.3446611 -0.611129042  1.0488088 -1.35148447
#> [20,]  0.6660343 -1.412520381 -0.9534626 -0.25223213
#> [21,] -0.4098672  0.361989013  1.0488088 -0.24348708
#> attr(,"center")
#>  [1]   0.5238095   2.5714286   5.7142857 215.8476190   3.6919048   3.7142857
#>  [7] 134.1428571  17.9676190   0.4761905   3.0584286
#> attr(,"scale")
#>  [1]   0.4994328   1.4333254   1.7766436 139.5388753   0.5403316   0.6998542
#>  [7]  61.3439062   1.7469617   0.4994328   1.1435045
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 44
#> 
#> $lambda.min
#> [1] 0.295508
#> 
#> $null.dev
#> [1] 44.99583
#> 
#> $Bias
#> [1] 7.435808
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 4.638194