Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 26.257776 25.699288 25.241531 23.815861 22.652343 20.618526 18.273090
#>   [8] 16.324591 14.715883 13.396246 12.322960 11.459234 10.772534 10.241982
#>  [15]  9.830994  9.427989  8.922384  8.424987  8.021786  7.426209  7.191487
#>  [22]  7.030456  6.915154  6.820805  6.750914  6.701514  6.669220  6.635902
#>  [29]  6.608023  8.377839  8.337535  8.337775  8.325197  8.833462  8.838554
#>  [36]  9.578830 10.355717 12.848317 13.216720 13.359283 13.597446 14.099061
#>  [43] 14.477903 14.987667 17.377251 17.544038 17.459803 17.607512 17.754965
#>  [50] 17.771429 17.742938 19.959183 19.964886 20.375589 20.652545 21.973310
#>  [57] 21.974853 21.976250 22.026615 22.152652 22.713331 23.707931 24.231830
#>  [64] 25.001867 24.410921 24.406388 24.405047 24.850791 24.851774 24.845834
#>  [71] 24.846719 24.906907 24.908469 25.134032 25.611668 25.528505 25.512303
#>  [78] 25.513490 25.514607 25.515660 25.516652 25.517588 25.518470 25.519292
#>  [85] 25.532133 25.532884 25.535718 25.536406 25.537060 25.537755 25.547022
#>  [92] 25.547764 25.548472 25.549149 25.502402 25.502742 25.503058 25.503358
#>  [99] 25.503890 25.508454
#> 
#> $cvse
#>   [1] 9.878534 9.892730 9.904470 9.446486 9.126729 8.388406 7.479184 6.726223
#>   [9] 6.099853 5.574527 5.129456 4.747878 4.416491 4.124157 3.845446 3.600424
#>  [17] 3.390558 3.192221 3.014293 2.858305 2.716015 2.585123 2.493063 2.447967
#>  [25] 2.406819 2.369371 2.335426 2.304789 2.277154 3.334047 3.319224 3.318050
#>  [33] 3.318359 3.374829 3.374830 3.393604 3.880539 4.444518 4.628152 4.628460
#>  [41] 4.866256 5.251021 5.540813 5.784451 6.024746 6.035096 6.057935 6.075304
#>  [49] 6.090456 6.088877 6.075159 7.825089 7.825688 7.800879 7.814085 7.975534
#>  [57] 7.976358 7.977120 7.979949 7.985824 8.053325 8.177133 8.151550 8.194142
#>  [65] 8.104119 8.104028 8.103524 8.097565 8.097165 8.054779 8.054407 8.045680
#>  [73] 8.045413 8.071279 8.136265 8.139463 8.134871 8.134825 8.134757 8.134670
#>  [81] 8.134565 8.134446 8.134313 8.134171 8.132230 8.132069 8.131941 8.131770
#>  [89] 8.131596 8.131413 8.131077 8.130900 8.130725 8.130551 8.087758 8.087309
#>  [97] 8.086861 8.086409 8.085944 8.084928
#> 
#> $fold
#>  [1]  8  7 10  7  8  9 10  5  1  5  1  1  4  3  9  2  4  3  2  6  6
#> 
#> $lambda
#>   [1] 4.205837194 3.922377843 3.658022705 3.411484217 3.181561599 2.967134995
#>   [7] 2.767160027 2.580662703 2.406734674 2.244528812 2.093255082 1.952176695
#>  [13] 1.820606520 1.697903734 1.583470705 1.476750079 1.377222066 1.284401908
#>  [19] 1.197837518 1.117107279 1.041817987 0.971602942 0.906120155 0.845050689
#>  [25] 0.788097101 0.734981994 0.685446668 0.639249856 0.596166556 0.555986925
#>  [31] 0.518515267 0.483569074 0.450978136 0.420583719 0.392237784 0.365802270
#>  [37] 0.341148422 0.318156160 0.296713500 0.276716003 0.258066270 0.240673467
#>  [43] 0.224452881 0.209325508 0.195217670 0.182060652 0.169790373 0.158347070
#>  [49] 0.147675008 0.137722207 0.128440191 0.119783752 0.111710728 0.104181799
#>  [55] 0.097160294 0.090612016 0.084505069 0.078809710 0.073498199 0.068544667
#>  [61] 0.063924986 0.059616656 0.055598693 0.051851527 0.048356907 0.045097814
#>  [67] 0.042058372 0.039223778 0.036580227 0.034114842 0.031815616 0.029671350
#>  [73] 0.027671600 0.025806627 0.024067347 0.022445288 0.020932551 0.019521767
#>  [79] 0.018206065 0.016979037 0.015834707 0.014767501 0.013772221 0.012844019
#>  [85] 0.011978375 0.011171073 0.010418180 0.009716029 0.009061202 0.008450507
#>  [91] 0.007880971 0.007349820 0.006854467 0.006392499 0.005961666 0.005559869
#>  [97] 0.005185153 0.004835691 0.004509781 0.004205837
#> 
#> $fit
#> $beta
#>               4.2058     3.9224    3.6580    3.4115    3.1816    2.9671
#> (Intercept) 19.58571 21.4412507 23.171730 24.785581 26.290664 27.694309
#> am           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000 -0.5701991 -1.101969 -1.597899 -2.060405 -2.491739
#>                2.7672    2.5807    2.4067    2.2445    2.0933    1.9522
#> (Intercept) 29.003353 30.224173 31.362712 32.424519 33.414763 34.338267
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.894004 -3.269157 -3.619025 -3.945314 -4.249612 -4.533402
#>                1.8206   1.6979       1.5835      1.4768     1.3772     1.2844
#> (Intercept) 35.199531 36.00275 36.680910614 36.04887627 35.3232936 34.0129327
#> am           0.000000  0.00000  0.000000000  0.00000000  0.0000000  0.0000000
#> carb         0.000000  0.00000  0.000000000  0.00000000  0.0000000  0.0000000
#> cyl          0.000000  0.00000  0.000000000  0.00000000  0.0000000  0.0000000
#> disp         0.000000  0.00000  0.000000000  0.00000000  0.0000000  0.0000000
#> drat         0.000000  0.00000  0.000000000  0.00000000  0.0000000  0.0000000
#> gear         0.000000  0.00000  0.000000000  0.00000000  0.0000000  0.0000000
#> hp           0.000000  0.00000  0.000000000  0.00000000  0.0000000  0.0000000
#> qsec         0.000000  0.00000  0.003735127  0.07381397  0.1408821  0.2114284
#> vs           0.000000  0.00000  0.000000000  0.00000000  0.0000000  0.0000000
#> wt          -4.798065 -5.04489 -5.273658025 -5.46164076 -5.6044561 -5.5865416
#>                 1.1978     1.1171     1.0418     0.9716     0.9061     0.8451
#> (Intercept) 32.7908589 31.6511487 30.5882512 29.5969894 28.6725354 27.8103865
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.2772217  0.3385807  0.3958043  0.4491712  0.4989414  0.5453572
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.5698342 -5.5542528 -5.5397215 -5.5261696 -5.5135311 -5.5017443
#>                 0.7881     0.7350     0.6854     0.6392     0.5962    0.5560
#> (Intercept) 27.0063435 26.2564903 25.5571747 24.9049907 24.2967616 23.729525
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> qsec         0.5886448  0.6290149  0.6666642  0.7017761  0.7345215  0.765060
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> wt          -5.4907519 -5.4805004 -5.4709398 -5.4620235 -5.4537082 -5.445953
#>                 0.5185     0.4836     0.4510     0.4206     0.3922     0.3658
#> (Intercept) 23.3001323 23.3000981 23.3000978 23.3000978 23.3000978 23.3000978
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.7881774  0.7881793  0.7881793  0.7881793  0.7881793  0.7881793
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.4400829 -5.4400825 -5.4400825 -5.4400825 -5.4400825 -5.4400825
#>                 0.3411     0.3182      0.2967      0.2767     0.2581     0.2407
#> (Intercept) 23.3000978 23.3000978 23.45016947 24.08869923 21.0780705 21.0765164
#> am           0.0000000  0.0000000  0.00000000  0.00000000  2.1813868  2.1818821
#> carb         0.0000000  0.0000000 -0.01341933 -0.07053726 -0.5611605 -0.5612005
#> cyl          0.0000000  0.0000000  0.00000000  0.00000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.00000000  0.00000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.00000000  0.00000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.00000000  0.00000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.00000000  0.00000000  0.0000000  0.0000000
#> qsec         0.7881793  0.7881793  0.78100745  0.75049393  0.7525811  0.7526237
#> vs           0.0000000  0.0000000  0.00000000  0.00000000  0.0000000  0.0000000
#> wt          -5.4400825 -5.4400825 -5.43432044 -5.40979185 -4.2847268 -4.2845011
#>                 0.2245     0.2093     0.1952      0.1821      0.1698
#> (Intercept) 21.0754686 21.0747622 28.0683385 28.06632366 28.06478862
#> am           2.1822159  2.1824408  0.0000000  0.00000000  0.00000000
#> carb        -0.5612274 -0.5612455 -0.8983655 -0.89852593 -0.89865187
#> cyl          0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> disp         0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> drat         0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> gear         0.0000000  0.0000000  2.0342019  2.03460292  2.03490913
#> hp           0.0000000  0.0000000 -0.0218705 -0.02187043 -0.02187029
#> qsec         0.7526523  0.7526717  0.0000000  0.00000000  0.00000000
#> vs           0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> wt          -4.2843489 -4.2842464 -3.0806186 -3.08031375 -3.08008266
#>                  0.1583     0.1477      0.1377      0.1284      0.1198
#> (Intercept) 28.06361554 28.0627171 28.06202787 28.06149856 28.06109173
#> am           0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> carb        -0.89875017 -0.8988266 -0.89888585 -0.89893170 -0.89896713
#> cyl          0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> drat         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> gear         2.03514351  2.0353232  2.03546121  2.03556724  2.03564877
#> hp          -0.02187014 -0.0218700 -0.02186988 -0.02186977 -0.02186969
#> qsec         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> wt          -3.07990671 -3.0797723 -3.07966940 -3.07959048 -3.07952988
#>                  0.1117     0.1042     0.0972      0.0906      0.0845
#> (Intercept) 23.51182180 23.8036181 24.0772151 24.31313381 24.31138611
#> am           1.22094052  1.2425625  1.2623920  1.28036583  1.28079992
#> carb        -0.80980953 -0.8039603 -0.7986000 -0.79373230 -0.79360607
#> cyl          0.00000000  0.0000000  0.0000000  0.00000000  0.00000000
#> disp         0.00000000  0.0000000  0.0000000  0.00000000  0.00000000
#> drat        -0.21786965 -0.3589633 -0.4906603 -0.60529755 -0.60521071
#> gear         1.56110992  1.5943789  1.6258292  1.65242741  1.65189981
#> hp          -0.01620651 -0.0162506 -0.0162961 -0.01632793 -0.01632243
#> qsec         0.32399353  0.3319821  0.3392946  0.34593029  0.34610902
#> vs           0.00000000  0.0000000  0.0000000  0.00000000  0.00000000
#> wt          -3.15408573 -3.1772958 -3.1986740 -3.21781035 -3.21816160
#>                  0.0788      0.0735      0.0685      0.0639      0.0596
#> (Intercept) 24.30986298 24.30856422 24.30747182 24.30656102 24.30580594
#> am           1.28115343  1.28144134  1.28167602  1.28186752  1.28202396
#> carb        -0.79350361 -0.79342067 -0.79335362 -0.79329944 -0.79325566
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat        -0.60510779 -0.60500541 -0.60491127 -0.60482826 -0.60475688
#> gear         1.65145444  1.65108257  1.65077430  1.65051998  1.65031083
#> hp          -0.01631778 -0.01631389 -0.01631066 -0.01630798 -0.01630577
#> qsec         0.34625637  0.34637748  0.34647683  0.34655821  0.34662481
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -3.21845109 -3.21868919 -3.21888463 -3.21904474 -3.21917570
#>                  0.0556      0.0519      0.0484      0.0451      0.0421
#> (Intercept) 24.30518230 24.30466855 24.30424608 22.94003944 22.93590799
#> am           1.28215187  1.28225653  1.28234219  1.13222560  1.13203108
#> carb        -0.79322025 -0.79319161 -0.79316841 -0.81858610 -0.81875917
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat        -0.60469645 -0.60464582 -0.60460370 -0.52734902 -0.52737960
#> gear         1.65013924  1.64999868  1.64988369  1.78014778  1.78059933
#> hp          -0.01630395 -0.01630246 -0.01630123 -0.01629567 -0.01629192
#> qsec         0.34667926  0.34672377  0.34676013  0.41191269  0.41208534
#> vs           0.00000000  0.00000000  0.00000000 -0.45082244 -0.45146527
#> wt          -3.21928266 -3.21936992 -3.21944104 -3.28989334 -3.28995737
#>                 0.0392      0.0366      0.0341      0.0318      0.0297
#> (Intercept) 22.9325541 22.92984041 22.92765031 22.92588634 22.92446783
#> am           1.1318756  1.13175053  1.13164942  1.13156738  1.13150061
#> carb        -0.8189025 -0.81902115 -0.81911935 -0.81920053 -0.81926757
#> cyl          0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat        -0.5274055 -0.52742780 -0.52744695 -0.52746339 -0.52747745
#> gear         1.7809719  1.78127921  1.78153267  1.78174165  1.78191392
#> hp          -0.0162889 -0.01628647 -0.01628453 -0.01628299 -0.01628176
#> qsec         0.4122236  0.41233407  0.41242197  0.41249176  0.41254705
#> vs          -0.4519826 -0.45239881 -0.45273360 -0.45300287 -0.45321944
#> wt          -3.2900020 -3.29003229 -3.29005216 -3.29006439 -3.29007107
#>                  0.0277      0.0258      0.0241      0.0224     0.0209
#> (Intercept) 22.92332860 22.92241470 22.92168237 22.92109616 22.9206275
#> am           1.13144613  1.13140157  1.13136503  1.13133499  1.1313102
#> carb        -0.81932288 -0.81936844 -0.81940593 -0.81943674 -0.8194620
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> drat        -0.52748944 -0.52749966 -0.52750836 -0.52751578 -0.5275221
#> gear         1.78205591  1.78217296  1.78226944  1.78234899  1.7824146
#> hp          -0.01628079 -0.01628003 -0.01627943 -0.01627896 -0.0162786
#> qsec         0.41259075  0.41262521  0.41265234  0.41267364  0.4126903
#> vs          -0.45339360 -0.45353365 -0.45364623 -0.45373670 -0.4538094
#> wt          -3.29007380 -3.29007378 -3.29007191 -3.29006886 -3.2900651
#>                  0.0195     0.0182      0.0170      0.0158      0.0148
#> (Intercept) 22.92025317 22.9199547 22.91971688 22.91952774 22.91937751
#> am           1.13128980  1.1312729  1.13125885  1.13124719  1.13123749
#> carb        -0.81948281 -0.8194998 -0.81951378 -0.81952520 -0.81953455
#> cyl          0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> drat        -0.52752757 -0.5275322 -0.52753626 -0.52753972 -0.52754270
#> gear         1.78246869  1.7825133  1.78255012  1.78258048  1.78260552
#> hp          -0.01627832 -0.0162781 -0.01627794 -0.01627781 -0.01627772
#> qsec         0.41270336  0.4127135  0.41272141  0.41272752  0.41273223
#> vs          -0.45386777 -0.4539146 -0.45395225 -0.45398242 -0.45400661
#> wt          -3.29006109 -3.2900570 -3.29005296 -3.29004916 -3.29004563
#>                  0.0138      0.0128      0.0120      0.0112     0.0104
#> (Intercept) 22.91925835 22.91916396 22.91908931 22.91903034 22.9189838
#> am           1.13122940  1.13122266  1.13121704  1.13121235  1.1312084
#> carb        -0.81954221 -0.81954846 -0.81955358 -0.81955776 -0.8195612
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> drat        -0.52754526 -0.52754746 -0.52754935 -0.52755097 -0.5275524
#> gear         1.78262617  1.78264321  1.78265725  1.78266883  1.7826784
#> hp          -0.01627764 -0.01627759 -0.01627755 -0.01627752 -0.0162775
#> qsec         0.41273586  0.41273863  0.41274075  0.41274235  0.4127436
#> vs          -0.45402601 -0.45404156 -0.45405402 -0.45406400 -0.4540720
#> wt          -3.29004241 -3.29003950 -3.29003691 -3.29003462 -3.2900326
#>                  0.0097     0.0091      0.0085      0.0079      0.0073
#> (Intercept) 24.05804516 24.0609324 24.06355545 24.06593849 24.06810351
#> am           1.05616059  1.0557410  1.05535884  1.05501094  1.05469428
#> carb        -0.83248106 -0.8325528 -0.83261827 -0.83267804 -0.83273257
#> cyl         -0.10167791 -0.1019436 -0.10218505 -0.10240433 -0.10260353
#> disp         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> drat        -0.56113395 -0.5611679 -0.56119918 -0.56122794 -0.56125435
#> gear         1.78603737  1.7862641  1.78647155  1.78666110  1.78683422
#> hp          -0.01567074 -0.0156715 -0.01567219 -0.01567283 -0.01567341
#> qsec         0.38016652  0.3800430  0.37993061  0.37982838  0.37973540
#> vs          -0.55989462 -0.5603199 -0.56070674 -0.56105861 -0.56137860
#> wt          -3.22647182 -3.2261882 -3.22593008 -3.22569514 -3.22548135
#>                  0.0069      0.0064      0.0060      0.0056      0.0052
#> (Intercept) 24.07007045 24.07185746 24.07348101 24.07495606 24.07629620
#> am           1.05440611  1.05414393  1.05390543  1.05368851  1.05349123
#> carb        -0.83278229 -0.83282761 -0.83286890 -0.83290651 -0.83294077
#> cyl         -0.10278448 -0.10294886 -0.10309818 -0.10323384 -0.10335707
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat        -0.56127859 -0.56130079 -0.56132113 -0.56133973 -0.56135674
#> gear         1.78699224  1.78713640  1.78726786  1.78738769  1.78749688
#> hp          -0.01567395 -0.01567444 -0.01567489 -0.01567529 -0.01567567
#> qsec         0.37965084  0.37957395  0.37950403  0.37944046  0.37938267
#> vs          -0.56166956 -0.56193410 -0.56217459 -0.56239321 -0.56259191
#> wt          -3.22528684 -3.22510989 -3.22494895 -3.22480258 -3.22466947
#>                  0.0048        0.0045        0.0042
#> (Intercept) 24.07751379  2.407743e+01 24.0193401831
#> am           1.05331185  1.053157e+00  1.0670014953
#> carb        -0.83297195 -8.330692e-01 -0.8026539224
#> cyl         -0.10346902 -1.037392e-01 -0.1045669711
#> disp         0.00000000  5.579821e-06  0.0007843061
#> drat        -0.56137228 -5.608192e-01 -0.5551520286
#> gear         1.78759634  1.787809e+00  1.7630346133
#> hp          -0.01567601 -1.568030e-02 -0.0162493082
#> qsec         0.37933013  3.792380e-01  0.3889596826
#> vs          -0.56277252 -5.630377e-01 -0.5214445923
#> wt          -3.22454844 -3.224400e+00 -3.2999975209
#> 
#> $iter
#>   [1]   0   3   3   3   2   2   2   2   2   2   2   2   2   2   4   3   3   3
#>  [19]   3   3   3   3   3   3   3   3   3   3   3   3   2   1   1   1   1   1
#>  [37]   1   1  10  12  41   1   1   1 109   1   1   1   1   1   1   1  63  17
#>  [55]  16  15   1   1   1   1   1   1   1   1   1  39   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1  44   1   1
#>  [91]   1   1   1   1   1   1   1   1   2  53
#> 
#> $lambda
#>   [1] 4.205837194 3.922377843 3.658022705 3.411484217 3.181561599 2.967134995
#>   [7] 2.767160027 2.580662703 2.406734674 2.244528812 2.093255082 1.952176695
#>  [13] 1.820606520 1.697903734 1.583470705 1.476750079 1.377222066 1.284401908
#>  [19] 1.197837518 1.117107279 1.041817987 0.971602942 0.906120155 0.845050689
#>  [25] 0.788097101 0.734981994 0.685446668 0.639249856 0.596166556 0.555986925
#>  [31] 0.518515267 0.483569074 0.450978136 0.420583719 0.392237784 0.365802270
#>  [37] 0.341148422 0.318156160 0.296713500 0.276716003 0.258066270 0.240673467
#>  [43] 0.224452881 0.209325508 0.195217670 0.182060652 0.169790373 0.158347070
#>  [49] 0.147675008 0.137722207 0.128440191 0.119783752 0.111710728 0.104181799
#>  [55] 0.097160294 0.090612016 0.084505069 0.078809710 0.073498199 0.068544667
#>  [61] 0.063924986 0.059616656 0.055598693 0.051851527 0.048356907 0.045097814
#>  [67] 0.042058372 0.039223778 0.036580227 0.034114842 0.031815616 0.029671350
#>  [73] 0.027671600 0.025806627 0.024067347 0.022445288 0.020932551 0.019521767
#>  [79] 0.018206065 0.016979037 0.015834707 0.014767501 0.013772221 0.012844019
#>  [85] 0.011978375 0.011171073 0.010418180 0.009716029 0.009061202 0.008450507
#>  [91] 0.007880971 0.007349820 0.006854467 0.006392499 0.005961666 0.005559869
#>  [97] 0.005185153 0.004835691 0.004509781 0.004205837
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 38
#> 
#> $loss
#>   [1] 523.42571 452.11463 392.45230 342.76239 301.59761 267.70919 240.02034
#>   [8] 217.60323 199.65896 185.50028 174.53650 166.26045 160.23715 156.09400
#>  [15] 153.02711 143.01673 135.11253 127.92076 121.96063 117.05196 113.03927
#>  [22] 109.78855 107.18443 105.12764 103.53287 102.32686 101.44676 100.83876
#>  [29] 100.45679 100.26151 100.21700 100.21700 100.21700 100.21700 100.21700
#>  [36] 100.21700 100.21700 100.21700  99.93599  98.87009  89.01693  89.01693
#>  [43]  89.01693  89.01693  82.82253  82.82253  82.82253  82.82253  82.82252
#>  [50]  82.82252  82.82252  82.82252  79.64134  79.46865  79.37705  79.35198
#>  [57]  79.35197  79.35196  79.35196  79.35195  79.35195  79.35195  79.35195
#>  [64]  79.35195  79.35195  79.13727  79.13726  79.13725  79.13725  79.13724
#>  [71]  79.13724  79.13724  79.13724  79.13724  79.13724  79.13724  79.13724
#>  [78]  79.13724  79.13724  79.13724  79.13724  79.13724  79.13724  79.13724
#>  [85]  79.13724  79.13724  79.13724  79.09919  79.09918  79.09917  79.09917
#>  [92]  79.09916  79.09916  79.09915  79.09915  79.09915  79.09915  79.09915
#>  [99]  79.09902  79.08972
#> 
#> $linear.predictors
#>      4.2058   3.9224   3.6580   3.4115   3.1816   2.9671   2.7672   2.5807
#> 1  19.58571 19.94733 20.28457 20.59909 20.89240 21.16595 21.42106 21.65898
#> 2  19.58571 19.80193 20.00357 20.19162 20.36700 20.53056 20.68309 20.82535
#> 3  19.58571 19.47977 19.38096 19.28881 19.20287 19.12273 19.04798 18.97827
#> 4  19.58571 19.46836 19.35892 19.25685 19.16166 19.07289 18.99010 18.91289
#> 5  19.58571 19.62232 19.65645 19.68828 19.71797 19.74566 19.77148 19.79556
#> 6  19.58571 19.64512 19.70053 19.75220 19.80039 19.84533 19.88724 19.92633
#> 7  19.58571 19.47977 19.38096 19.28881 19.20287 19.12273 19.04798 18.97827
#> 8  19.58571 19.47977 19.38096 19.28881 19.20287 19.12273 19.04798 18.97827
#> 9  19.58571 19.12054 18.68672 18.28213 17.90482 17.55293 17.22476 16.91871
#> 10 19.58571 19.31441 19.06139 18.82542 18.60535 18.40012 18.20872 18.03022
#> 11 19.58571 18.44771 17.38639 16.39661 15.47354 14.61268 13.80983 13.06110
#> 12 19.58571 20.18681 20.74740 21.27020 21.75777 22.21248 22.63655 23.03203
#> 13 19.58571 20.03571 20.45538 20.84676 21.21177 21.55217 21.86963 22.16570
#> 14 19.58571 19.43415 19.29280 19.16098 19.03804 18.92339 18.81646 18.71674
#> 15 19.58571 19.25169 18.94017 18.64965 18.37871 18.12603 17.89038 17.67061
#> 16 19.58571 19.24884 18.93466 18.64166 18.36841 18.11357 17.87591 17.65427
#> 17 19.58571 20.57854 21.50445 22.36796 23.17327 23.92431 24.62473 25.27794
#> 18 19.58571 19.63372 19.67849 19.72024 19.75918 19.79550 19.82936 19.86095
#> 19 19.58571 19.86180 20.11928 20.35940 20.58334 20.79219 20.98696 21.16861
#> 20 19.58571 19.40564 19.23770 19.08108 18.93502 18.79880 18.67176 18.55328
#> 21 19.58571 19.85610 20.10826 20.34342 20.56274 20.76727 20.95802 21.13592
#>      2.4067   2.2445   2.0933   1.9522   1.8206    1.6979    1.5835    1.4768
#> 1  21.88087 22.08780 22.28078 22.46075 22.62860 22.785136 22.925407 22.954355
#> 2  20.95801 21.08174 21.19713 21.30474 21.40509 21.498689 21.582716 21.602973
#> 3  18.91326 18.85264 18.79610 18.74337 18.69419 18.648326 18.603099 18.517146
#> 4  18.84088 18.77373 18.71110 18.65270 18.59823 18.547428 18.509578 18.644118
#> 5  19.81802 19.83897 19.85850 19.87672 19.89370 19.909548 19.932644 20.102522
#> 6  19.96278 19.99678 20.02848 20.05805 20.08563 20.111344 20.154422 20.535048
#> 7  18.91326 18.85264 18.79610 18.74337 18.69419 18.648326 18.607880 18.611628
#> 8  18.91326 18.85264 18.79610 18.74337 18.69419 18.648326 18.610121 18.655916
#> 9  16.63328 16.36709 16.11884 15.88732 15.67141 15.470045 15.282114 15.104361
#> 10 17.86375 17.70850 17.56371 17.42868 17.30275 17.185307 17.075904 16.976082
#> 11 12.36283 11.71162 11.10430 10.53791 10.00969  9.517074  9.061364  8.702437
#> 12 23.40086 23.74483 24.06562 24.36478 24.64379 24.903990 25.151586 25.470425
#> 13 22.44181 22.69932 22.93947 23.16343 23.37230 23.567094 23.756083 24.062949
#> 14 18.62374 18.53701 18.45613 18.38069 18.31034 18.244734 18.180646 18.069142
#> 15 17.46565 17.27451 17.09625 16.93000 16.77496 16.630369 16.487622 16.213649
#> 16 17.44756 17.25478 17.07500 16.90734 16.75097 16.605145 16.467379 16.307396
#> 17 25.88713 26.45526 26.98510 27.47923 27.94006 28.369829 28.764990 29.032870
#> 18 19.89040 19.91787 19.94349 19.96738 19.98967 20.010446 20.017574 19.805778
#> 19 21.33801 21.49600 21.64334 21.78074 21.90889 22.028402 22.130772 22.064248
#> 20 18.44279 18.33975 18.24365 18.15402 18.07044 17.992490 17.908484 17.628503
#> 21 21.30182 21.45654 21.60084 21.73541 21.86091 21.977953 22.089615 22.238455
#>       1.3772    1.2844    1.1978    1.1171    1.0418    0.9716    0.9061
#> 1  22.958537 22.856306 22.760962 22.672044 22.589119 22.511783 22.439659
#> 2  21.608295 21.550138 21.495899 21.445315 21.398141 21.354146 21.313116
#> 3  18.441777 18.393742 18.348942 18.307162 18.268198 18.231860 18.197971
#> 4  18.780511 18.958582 19.124655 19.279535 19.423977 19.558684 19.684313
#> 5  20.262720 20.420434 20.567521 20.704696 20.832625 20.951933 21.063199
#> 6  20.895456 21.257038 21.594258 21.908750 22.202046 22.475576 22.730670
#> 7  18.622106 18.664370 18.703786 18.740546 18.774828 18.806799 18.836616
#> 8  18.706636 18.791227 18.870119 18.943694 19.012310 19.076302 19.135981
#> 9  14.964505 14.954563 14.945291 14.936644 14.928579 14.921058 14.914044
#> 10 16.898197 16.896273 16.894479 16.892806 16.891245 16.889790 16.888433
#> 11  8.432959  8.485073  8.533675  8.579002  8.621274  8.660697  8.697463
#> 12 25.736464 25.839053 25.934730 26.023958 26.107173 26.184780 26.257156
#> 13 24.327359 24.472791 24.608423 24.734915 24.852881 24.962897 25.065498
#> 14 17.972289 17.915104 17.861772 17.812035 17.765650 17.722391 17.682047
#> 15 15.973175 15.818725 15.674682 15.540346 15.415065 15.298226 15.189263
#> 16 16.176199 16.137535 16.101476 16.067847 16.036485 16.007236 15.979959
#> 17 29.224658 29.133636 29.048746 28.969578 28.895745 28.826888 28.762672
#> 18 19.599958 19.369308 19.154199 18.953587 18.766496 18.592014 18.429292
#> 19 21.982622 21.815353 21.659354 21.513869 21.378189 21.251653 21.133646
#> 20 17.372264 17.155834 16.953987 16.765744 16.590188 16.426464 16.273774
#> 21 22.363312 22.414916 22.463043 22.507927 22.549785 22.588822 22.625229
#>       0.8451    0.7881    0.7350    0.6854    0.6392    0.5962   0.5560
#> 1  22.372396 22.309666 22.251164 22.196605 22.145723 22.098270 22.05402
#> 2  21.274851 21.239166 21.205885 21.174848 21.145902 21.118907 21.09373
#> 3  18.166366 18.136891 18.109402 18.083767 18.059858 18.037562 18.01677
#> 4  19.801474 19.910739 20.012640 20.107673 20.196301 20.278956 20.35604
#> 5  21.166966 21.263740 21.353992 21.438161 21.516657 21.589863 21.65813
#> 6  22.968572 23.190440 23.397355 23.590325 23.770289 23.938124 24.09465
#> 7  18.864423 18.890356 18.914541 18.937097 18.958132 18.977749 18.99604
#> 8  19.191637 19.243543 19.291950 19.337095 19.379197 19.418462 19.45508
#> 9  14.907503 14.901402 14.895713 14.890407 14.885459 14.880844 14.87654
#> 10 16.887167 16.885987 16.884886 16.883859 16.882902 16.882009 16.88118
#> 11  8.731752  8.763729  8.793551  8.821363  8.847301  8.871491  8.89405
#> 12 26.324654 26.387603 26.446309 26.501059 26.552119 26.599738 26.64415
#> 13 25.161185 25.250422 25.333645 25.411259 25.483642 25.551147 25.61410
#> 14 17.644423 17.609334 17.576610 17.546092 17.517630 17.491087 17.46633
#> 15 15.087643 14.992872 14.904488 14.822061 14.745190 14.673499 14.60664
#> 16 15.954520 15.930796 15.908670 15.888036 15.868792 15.850846 15.83411
#> 17 28.702784 28.646932 28.594845 28.546268 28.500965 28.458715 28.41931
#> 18 18.277537 18.136009 18.004020 17.880926 17.766129 17.659069 17.55922
#> 19 21.023592 20.920955 20.825235 20.735967 20.652715 20.575074 20.50267
#> 20 16.131375 15.998573 15.874721 15.759217 15.651497 15.551038 15.45735
#> 21 22.659181 22.690846 22.720376 22.747916 22.773600 22.797553 22.81989
#>       0.5185    0.4836    0.4510    0.4206    0.3922    0.3658    0.3411
#> 1  22.020515 22.020513 22.020513 22.020513 22.020513 22.020513 22.020513
#> 2  21.074674 21.074672 21.074672 21.074672 21.074672 21.074672 21.074672
#> 3  18.001027 18.001026 18.001026 18.001026 18.001026 18.001026 18.001026
#> 4  20.414393 20.414398 20.414398 20.414398 20.414398 20.414398 20.414398
#> 5  21.709816 21.709820 21.709820 21.709820 21.709820 21.709820 21.709820
#> 6  24.213134 24.213144 24.213144 24.213144 24.213144 24.213144 24.213144
#> 7  19.009894 19.009895 19.009895 19.009895 19.009895 19.009895 19.009895
#> 8  19.482800 19.482803 19.482803 19.482803 19.482803 19.482803 19.482803
#> 9  14.873282 14.873282 14.873282 14.873282 14.873282 14.873282 14.873282
#> 10 16.880546 16.880546 16.880546 16.880546 16.880546 16.880546 16.880546
#> 11  8.911127  8.911128  8.911128  8.911128  8.911128  8.911128  8.911128
#> 12 26.677764 26.677767 26.677767 26.677767 26.677767 26.677767 26.677767
#> 13 25.661758 25.661762 25.661762 25.661762 25.661762 25.661762 25.661762
#> 14 17.447594 17.447592 17.447592 17.447592 17.447592 17.447592 17.447592
#> 15 14.556028 14.556024 14.556024 14.556024 14.556024 14.556024 14.556024
#> 16 15.821439 15.821438 15.821438 15.821438 15.821438 15.821438 15.821438
#> 17 28.389485 28.389483 28.389483 28.389483 28.389483 28.389483 28.389483
#> 18 17.483642 17.483636 17.483636 17.483636 17.483636 17.483636 17.483636
#> 19 20.447853 20.447848 20.447848 20.447848 20.447848 20.447848 20.447848
#> 20 15.386427 15.386421 15.386421 15.386421 15.386421 15.386421 15.386421
#> 21 22.836802 22.836803 22.836803 22.836803 22.836803 22.836803 22.836803
#>       0.3182    0.2967    0.2767    0.2581    0.2407    0.2245    0.2093
#> 1  22.020513 22.013955 21.986026 22.176317 22.176389 22.176438 22.176471
#> 2  21.074672 21.065568 21.026805 21.505157 21.505311 21.505414 21.505484
#> 3  18.001026 18.022015 18.111347 18.025220 18.025086 18.024996 18.024935
#> 4  20.414398 20.425972 20.475269 20.908946 20.908993 20.909024 20.909045
#> 5  21.709820 21.707998 21.700267 21.339094 21.339030 21.338987 21.338958
#> 6  24.213144 24.190292 24.093091 23.692968 23.693019 23.693053 23.693076
#> 7  19.009895 18.994866 18.930905 17.866203 17.866044 17.865936 17.865864
#> 8  19.482803 19.463471 19.381202 18.317752 18.317618 18.317528 18.317467
#> 9  14.873282 14.881757 14.917829 15.050663 15.050647 15.050637 15.050630
#> 10 16.880546 16.885627 16.907257 16.657986 16.657902 16.657846 16.657808
#> 11  8.911128  8.908824  8.899024  9.870022  9.870257  9.870416  9.870523
#> 12 26.677767 26.687460 26.728737 27.924653 27.924878 27.925030 27.925133
#> 13 25.661762 25.669109 25.700409 25.014207 25.014020 25.013894 25.013809
#> 14 17.447592 17.470119 17.565990 17.569555 17.569433 17.569350 17.569295
#> 15 14.556024 14.564027 14.598061 13.977353 13.977161 13.977032 13.976944
#> 16 15.821438 15.844546 15.942897 16.312483 16.312442 16.312414 16.312396
#> 17 28.389483 28.400230 28.445957 28.372966 28.372887 28.372834 28.372798
#> 18 17.483636 17.494304 17.539672 18.344658 18.344771 18.344848 18.344899
#> 19 20.447848 20.422201 20.313008 19.688809 19.688794 19.688785 19.688778
#> 20 15.386421 15.345000 15.168656 14.461383 14.461431 14.461464 14.461486
#> 21 22.836803 22.842659 22.867591 24.223605 24.223885 24.224073 24.224200
#>       0.1952    0.1821    0.1698    0.1583    0.1477    0.1377   0.1284
#> 1  22.134708 22.134462 22.134269 22.134118 22.134000 22.133909 22.13384
#> 2  21.349151 21.348982 21.348848 21.348742 21.348658 21.348593 21.34854
#> 3  17.949548 17.949476 17.949427 17.949392 17.949367 17.949349 17.94934
#> 4  20.317236 20.317326 20.317398 20.317454 20.317498 20.317533 20.31756
#> 5  23.225271 23.225516 23.225700 23.225838 23.225943 23.226023 23.22608
#> 6  22.626769 22.627004 22.627183 22.627320 22.627424 22.627504 22.62757
#> 7  19.324285 19.324290 19.324287 19.324283 19.324277 19.324272 19.32427
#> 8  19.324285 19.324290 19.324287 19.324283 19.324277 19.324272 19.32427
#> 9  15.001040 15.001000 15.000972 15.000950 15.000934 15.000922 15.00091
#> 10 16.048450 16.048307 16.048200 16.048118 16.048057 16.048009 16.04797
#> 11  9.920782  9.920944  9.921065  9.921156  9.921226  9.921279  9.92132
#> 12 27.085967 27.086071 27.086152 27.086215 27.086264 27.086302 27.08633
#> 13 23.557415 23.557201 23.557042 23.556922 23.556832 23.556763 23.55671
#> 14 18.249861 18.249812 18.249778 18.249753 18.249735 18.249722 18.24971
#> 15 13.389634 13.389369 13.389170 13.389019 13.388905 13.388818 13.38875
#> 16 16.701897 16.701949 16.701993 16.702030 16.702059 16.702083 16.70210
#> 17 29.310275 29.310413 29.310523 29.310608 29.310675 29.310726 29.31077
#> 18 19.106513 19.106847 19.107108 19.107311 19.107469 19.107591 19.10769
#> 19 20.488504 20.488389 20.488293 20.488216 20.488155 20.488106 20.48807
#> 20 12.727998 12.727817 12.727677 12.727568 12.727484 12.727419 12.72737
#> 21 23.460411 23.460535 23.460630 23.460703 23.460760 23.460804 23.46084
#>       0.1198   0.1117   0.1042   0.0972   0.0906   0.0845   0.0788   0.0735
#> 1  22.133784 22.17479 22.16024 22.14664 22.13490 22.13494 22.13498 22.13501
#> 2  21.348504 21.55193 21.53594 21.52098 21.50808 21.50813 21.50818 21.50822
#> 3  17.949325 17.71742 17.72468 17.73146 17.73734 17.73733 17.73733 17.73733
#> 4  20.317580 20.72035 20.80498 20.88391 20.95272 20.95273 20.95272 20.95271
#> 5  23.226131 22.74624 22.74517 22.74456 22.74329 22.74280 22.74239 22.74205
#> 6  22.627613 23.22706 23.21618 23.20584 23.19716 23.19740 23.19761 23.19779
#> 7  19.324263 18.74860 18.70471 18.66397 18.62807 18.62780 18.62758 18.62740
#> 8  19.324263 18.94300 18.90390 18.86755 18.83563 18.83546 18.83533 18.83523
#> 9  15.000906 14.98005 14.99264 15.00440 15.01463 15.01461 15.01460 15.01459
#> 10 16.047946 16.11724 16.13932 16.15981 16.17787 16.17801 16.17812 16.17822
#> 11  9.921351 10.26168 10.27601 10.28945 10.30106 10.30099 10.30092 10.30086
#> 12 27.086354 27.57802 27.55627 27.53587 27.51830 27.51843 27.51854 27.51863
#> 13 23.556670 23.71548 23.68925 23.66449 23.64344 23.64380 23.64411 23.64437
#> 14 18.249705 17.90663 17.96696 18.02343 18.07229 18.07206 18.07185 18.07167
#> 15 13.388700 13.05372 12.96561 12.88322 12.81178 12.81203 12.81226 12.81247
#> 16 16.702115 16.46499 16.47296 16.48052 16.48688 16.48673 16.48660 16.48650
#> 17 29.310797 28.96935 29.02379 29.07479 29.11882 29.11856 29.11833 29.11814
#> 18 19.107759 18.80060 18.73896 18.68157 18.63131 18.63116 18.63105 18.63097
#> 19 20.488039 20.33971 20.39562 20.44788 20.49326 20.49314 20.49303 20.49293
#> 20 12.727330 13.32962 13.37570 13.41826 13.45617 13.45673 13.45719 13.45757
#> 21 23.460863 23.95355 23.91110 23.87140 23.83702 23.83715 23.83727 23.83737
#>      0.0685   0.0639   0.0596   0.0556   0.0519   0.0484   0.0451   0.0421
#> 1  22.13503 22.13505 22.13507 22.13508 22.13509 22.13510 22.22989 22.22964
#> 2  21.50825 21.50827 21.50829 21.50830 21.50831 21.50832 21.62164 21.62147
#> 3  17.73732 17.73732 17.73732 17.73732 17.73732 17.73732 17.82394 17.82410
#> 4  20.95269 20.95267 20.95265 20.95264 20.95263 20.95262 20.79039 20.79038
#> 5  22.74177 22.74154 22.74135 22.74120 22.74107 22.74097 22.75988 22.75993
#> 6  23.19793 23.19805 23.19815 23.19823 23.19829 23.19835 23.42698 23.42765
#> 7  18.62725 18.62714 18.62704 18.62696 18.62690 18.62684 18.48466 18.48428
#> 8  18.83514 18.83507 18.83501 18.83497 18.83493 18.83490 18.73180 18.73153
#> 9  15.01457 15.01456 15.01456 15.01455 15.01454 15.01454 15.04996 15.04999
#> 10 16.17829 16.17835 16.17840 16.17844 16.17847 16.17850 16.25090 16.25099
#> 11 10.30081 10.30077 10.30074 10.30071 10.30068 10.30066 10.25464 10.25462
#> 12 27.51871 27.51877 27.51883 27.51887 27.51891 27.51894 27.47852 27.47853
#> 13 23.64459 23.64477 23.64492 23.64504 23.64514 23.64522 23.61199 23.61194
#> 14 18.07152 18.07140 18.07129 18.07121 18.07114 18.07108 18.11202 18.11206
#> 15 12.81264 12.81279 12.81291 12.81302 12.81310 12.81317 12.76107 12.76082
#> 16 16.48641 16.48634 16.48629 16.48624 16.48621 16.48617 16.54080 16.54094
#> 17 29.11798 29.11784 29.11773 29.11764 29.11757 29.11751 29.03921 29.03928
#> 18 18.63091 18.63087 18.63084 18.63081 18.63079 18.63077 18.71496 18.71536
#> 19 20.49283 20.49275 20.49269 20.49263 20.49258 20.49255 20.57238 20.57232
#> 20 13.45788 13.45813 13.45833 13.45850 13.45863 13.45874 13.36746 13.36744
#> 21 23.83746 23.83753 23.83759 23.83765 23.83769 23.83772 23.67690 23.67671
#>      0.0392   0.0366   0.0341   0.0318   0.0297   0.0277   0.0258   0.0241
#> 1  22.22944 22.22928 22.22914 22.22903 22.22894 22.22886 22.22880 22.22875
#> 2  21.62134 21.62123 21.62113 21.62106 21.62099 21.62094 21.62090 21.62087
#> 3  17.82422 17.82432 17.82440 17.82447 17.82452 17.82456 17.82460 17.82462
#> 4  20.79037 20.79036 20.79036 20.79036 20.79036 20.79036 20.79036 20.79036
#> 5  22.75998 22.76002 22.76006 22.76009 22.76012 22.76015 22.76017 22.76019
#> 6  23.42819 23.42863 23.42898 23.42927 23.42949 23.42968 23.42982 23.42994
#> 7  18.48397 18.48372 18.48352 18.48336 18.48323 18.48313 18.48304 18.48298
#> 8  18.73131 18.73112 18.73098 18.73086 18.73076 18.73068 18.73062 18.73057
#> 9  15.05001 15.05003 15.05004 15.05005 15.05006 15.05007 15.05008 15.05008
#> 10 16.25105 16.25111 16.25114 16.25117 16.25120 16.25122 16.25123 16.25124
#> 11 10.25461 10.25460 10.25460 10.25459 10.25459 10.25459 10.25459 10.25460
#> 12 27.47854 27.47855 27.47856 27.47856 27.47857 27.47857 27.47858 27.47858
#> 13 23.61190 23.61186 23.61183 23.61180 23.61177 23.61174 23.61172 23.61170
#> 14 18.11210 18.11213 18.11215 18.11217 18.11219 18.11221 18.11222 18.11223
#> 15 12.76062 12.76045 12.76031 12.76019 12.76009 12.76002 12.75995 12.75990
#> 16 16.54105 16.54115 16.54122 16.54129 16.54134 16.54138 16.54142 16.54145
#> 17 29.03934 29.03940 29.03945 29.03949 29.03952 29.03955 29.03958 29.03960
#> 18 18.71570 18.71597 18.71620 18.71638 18.71654 18.71666 18.71676 18.71685
#> 19 20.57227 20.57223 20.57219 20.57216 20.57214 20.57212 20.57210 20.57209
#> 20 13.36743 13.36742 13.36740 13.36739 13.36737 13.36736 13.36734 13.36733
#> 21 23.67656 23.67644 23.67634 23.67626 23.67620 23.67615 23.67612 23.67608
#>      0.0224   0.0209   0.0195   0.0182   0.0170   0.0158   0.0148   0.0138
#> 1  22.22871 22.22868 22.22865 22.22863 22.22861 22.22860 22.22859 22.22858
#> 2  21.62084 21.62082 21.62080 21.62078 21.62077 21.62076 21.62076 21.62075
#> 3  17.82465 17.82466 17.82468 17.82469 17.82470 17.82471 17.82471 17.82472
#> 4  20.79036 20.79036 20.79036 20.79036 20.79036 20.79036 20.79036 20.79036
#> 5  22.76021 22.76022 22.76023 22.76025 22.76026 22.76026 22.76027 22.76028
#> 6  23.43003 23.43010 23.43016 23.43021 23.43024 23.43027 23.43030 23.43032
#> 7  18.48293 18.48288 18.48285 18.48282 18.48280 18.48278 18.48277 18.48276
#> 8  18.73053 18.73050 18.73047 18.73045 18.73043 18.73042 18.73041 18.73040
#> 9  15.05009 15.05009 15.05009 15.05010 15.05010 15.05010 15.05010 15.05010
#> 10 16.25125 16.25125 16.25126 16.25126 16.25126 16.25126 16.25126 16.25126
#> 11 10.25460 10.25460 10.25460 10.25460 10.25461 10.25461 10.25461 10.25461
#> 12 27.47858 27.47858 27.47859 27.47859 27.47859 27.47859 27.47859 27.47859
#> 13 23.61168 23.61167 23.61165 23.61164 23.61163 23.61162 23.61162 23.61161
#> 14 18.11224 18.11225 18.11226 18.11227 18.11227 18.11228 18.11228 18.11229
#> 15 12.75985 12.75982 12.75979 12.75976 12.75974 12.75972 12.75971 12.75970
#> 16 16.54147 16.54149 16.54151 16.54152 16.54154 16.54154 16.54155 16.54156
#> 17 29.03962 29.03964 29.03965 29.03966 29.03967 29.03968 29.03968 29.03969
#> 18 18.71692 18.71697 18.71702 18.71706 18.71709 18.71711 18.71713 18.71715
#> 19 20.57208 20.57207 20.57207 20.57206 20.57206 20.57206 20.57205 20.57205
#> 20 13.36731 13.36730 13.36729 13.36728 13.36727 13.36726 13.36726 13.36725
#> 21 23.67606 23.67604 23.67602 23.67601 23.67600 23.67599 23.67598 23.67598
#>      0.0128   0.0120   0.0112   0.0104   0.0097   0.0091   0.0085   0.0079
#> 1  22.22857 22.22856 22.22856 22.22855 22.21034 22.21033 22.21032 22.21031
#> 2  21.62074 21.62074 21.62074 21.62073 21.60049 21.60048 21.60047 21.60046
#> 3  17.82472 17.82472 17.82472 17.82473 17.79919 17.79912 17.79906 17.79901
#> 4  20.79036 20.79036 20.79036 20.79037 20.74293 20.74272 20.74252 20.74235
#> 5  22.76028 22.76029 22.76029 22.76030 22.83934 22.83977 22.84015 22.84050
#> 6  23.43033 23.43034 23.43035 23.43036 23.42469 23.42471 23.42473 23.42475
#> 7  18.48275 18.48275 18.48274 18.48274 18.43315 18.43312 18.43310 18.43308
#> 8  18.73040 18.73039 18.73039 18.73038 18.66125 18.66115 18.66106 18.66098
#> 9  15.05010 15.05010 15.05010 15.05010 15.04503 15.04502 15.04502 15.04501
#> 10 16.25126 16.25126 16.25126 16.25126 16.21807 16.21794 16.21782 16.21771
#> 11 10.25461 10.25461 10.25461 10.25462 10.31260 10.31277 10.31292 10.31306
#> 12 27.47859 27.47859 27.47859 27.47859 27.43918 27.43902 27.43888 27.43875
#> 13 23.61160 23.61160 23.61160 23.61159 23.67469 23.67473 23.67476 23.67479
#> 14 18.11229 18.11229 18.11229 18.11230 18.09466 18.09467 18.09467 18.09468
#> 15 12.75969 12.75968 12.75967 12.75967 12.80916 12.80919 12.80922 12.80924
#> 16 16.54156 16.54157 16.54157 16.54157 16.54316 16.54320 16.54324 16.54328
#> 17 29.03970 29.03970 29.03970 29.03971 29.06972 29.06982 29.06991 29.06999
#> 18 18.71716 18.71717 18.71718 18.71719 18.68048 18.68044 18.68040 18.68036
#> 19 20.57205 20.57205 20.57205 20.57205 20.62101 20.62120 20.62138 20.62153
#> 20 13.36725 13.36724 13.36724 13.36724 13.36694 13.36667 13.36643 13.36621
#> 21 23.67597 23.67597 23.67597 23.67596 23.71392 23.71393 23.71394 23.71395
#>      0.0073   0.0069   0.0064   0.0060   0.0056   0.0052   0.0048   0.0045
#> 1  22.21030 22.21029 22.21028 22.21027 22.21027 22.21026 22.21025 22.21031
#> 2  21.60045 21.60045 21.60044 21.60043 21.60043 21.60042 21.60042 21.60046
#> 3  17.79895 17.79891 17.79887 17.79883 17.79879 17.79876 17.79873 17.79888
#> 4  20.74219 20.74204 20.74191 20.74179 20.74168 20.74158 20.74149 20.74105
#> 5  22.84082 22.84112 22.84138 22.84162 22.84184 22.84204 22.84222 22.84268
#> 6  23.42477 23.42478 23.42479 23.42481 23.42482 23.42483 23.42484 23.42498
#> 7  18.43307 18.43305 18.43304 18.43302 18.43301 18.43300 18.43299 18.43289
#> 8  18.66091 18.66084 18.66078 18.66073 18.66068 18.66063 18.66059 18.66043
#> 9  15.04500 15.04500 15.04499 15.04499 15.04498 15.04498 15.04497 15.04455
#> 10 16.21761 16.21752 16.21744 16.21737 16.21730 16.21724 16.21719 16.21669
#> 11 10.31319 10.31330 10.31341 10.31350 10.31359 10.31367 10.31374 10.31424
#> 12 27.43864 27.43853 27.43843 27.43834 27.43826 27.43819 27.43813 27.43825
#> 13 23.67481 23.67483 23.67485 23.67487 23.67489 23.67490 23.67492 23.67486
#> 14 18.09468 18.09469 18.09469 18.09469 18.09470 18.09470 18.09471 18.09454
#> 15 12.80926 12.80927 12.80929 12.80930 12.80931 12.80932 12.80933 12.80946
#> 16 16.54331 16.54335 16.54338 16.54340 16.54343 16.54345 16.54347 16.54386
#> 17 29.07006 29.07013 29.07019 29.07025 29.07030 29.07035 29.07039 29.07048
#> 18 18.68032 18.68029 18.68027 18.68024 18.68022 18.68020 18.68018 18.68075
#> 19 20.62168 20.62181 20.62193 20.62204 20.62214 20.62223 20.62231 20.62198
#> 20 13.36601 13.36583 13.36567 13.36552 13.36538 13.36525 13.36514 13.36442
#> 21 23.71396 23.71397 23.71397 23.71398 23.71398 23.71399 23.71399 23.71425
#>      0.0042
#> 1  22.22972
#> 2  21.60604
#> 3  17.82470
#> 4  20.74179
#> 5  22.83775
#> 6  23.42919
#> 7  18.43458
#> 8  18.66795
#> 9  14.98797
#> 10 16.18777
#> 11 10.34229
#> 12 27.43341
#> 13 23.67862
#> 14 18.09215
#> 15 12.80588
#> 16 16.57010
#> 17 29.08251
#> 18 18.67602
#> 19 20.60652
#> 20 13.36889
#> 21 23.69617
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 18.7 18.1 24.4 22.8 19.2 17.8 16.4 17.3 10.4 32.4 21.5 15.5 13.3
#> [16] 19.2 30.4 15.8 19.7 15.0 21.4
#> 
#> $X
#>               am        carb        cyl        disp        drat       gear
#>  [1,]  1.2747549  0.53611602 -0.1714986 -0.63555017  0.74318589  0.3175003
#>  [2,]  1.2747549  0.53611602 -0.1714986 -0.63555017  0.74318589  0.3175003
#>  [3,] -0.7844645 -0.64898255  1.0289915  1.17882634 -0.91477131 -1.0160010
#>  [4,] -0.7844645 -1.24153183 -0.1714986 -0.04587781 -1.77690905 -1.0160010
#>  [5,] -0.7844645 -0.64898255 -1.3719887 -0.75620621  0.27895788  0.3175003
#>  [6,] -0.7844645 -0.64898255 -1.3719887 -0.80973031  0.78739809  0.3175003
#>  [7,] -0.7844645  0.53611602 -0.1714986 -0.56660386  0.78739809  0.3175003
#>  [8,] -0.7844645  0.53611602 -0.1714986 -0.56660386  0.78739809  0.3175003
#>  [9,] -0.7844645 -0.05643326  1.0289915  0.41497383 -1.09162007 -1.0160010
#> [10,] -0.7844645 -0.05643326  1.0289915  0.41497383 -1.09162007 -1.0160010
#> [11,] -0.7844645  0.53611602  1.0289915  2.19487718 -1.40110542 -1.0160010
#> [12,]  1.2747549 -1.24153183 -1.3719887 -1.37309422  1.14109562  0.3175003
#> [13,] -0.7844645 -1.24153183 -1.3719887 -0.99751828  0.30106397 -1.0160010
#> [14,] -0.7844645 -0.64898255  1.0289915  0.79780727 -1.77690905 -1.0160010
#> [15,] -0.7844645  0.53611602  1.0289915  1.08810751  0.36738226 -1.0160010
#> [16,] -0.7844645 -0.64898255  1.0289915  1.54170164 -1.06951398 -1.0160010
#> [17,]  1.2747549 -0.64898255 -1.3719887 -1.22431535  0.45580665  1.6510017
#> [18,]  1.2747549  0.53611602  1.0289915  1.09717939  1.45058096  1.6510017
#> [19,]  1.2747549  1.72121458 -0.1714986 -0.77162841  0.12421521  1.6510017
#> [20,]  1.2747549  2.90631314  1.0289915  0.64358527 -0.05263356  1.6510017
#> [21,]  1.2747549 -0.64898255 -1.3719887 -0.98935359  1.20741391  0.3175003
#>                hp        qsec         vs          wt
#>  [1,] -0.63399790 -0.64390535 -0.8660254 -0.85048015
#>  [2,] -0.63399790 -0.36396720 -0.8660254 -0.50851280
#>  [3,]  0.34165840 -0.36396720 -0.8660254  0.24917957
#>  [4,] -0.70904839  1.23567936  1.1547005  0.27600054
#>  [5,] -1.35448256  1.12570366  1.1547005 -0.08608254
#>  [6,] -0.85914936  2.57538335  1.1547005 -0.13972448
#>  [7,] -0.43886664  0.27589142  1.1547005  0.24917957
#>  [8,] -0.43886664  0.57582515  1.1547005  0.24917957
#>  [9,]  0.41670888 -0.17400917 -0.8660254  1.09404010
#> [10,]  0.41670888 -0.07403126 -0.8660254  0.63808362
#> [11,]  0.79196130  0.11592677 -0.8660254  2.67647726
#> [12,] -1.29444217  0.86076219  1.1547005 -1.41372050
#> [13,] -0.82912916  1.13070255  1.1547005 -1.05834266
#> [14,] -0.03359403 -0.43895063 -0.8660254  0.35646345
#> [15,]  1.39236518 -1.16878937 -0.8660254  0.78559895
#> [16,]  0.34165840 -0.34897051 -0.8660254  0.79230419
#> [17,] -0.58896761 -0.42395395  1.1547005 -2.33502079
#> [18,]  1.67755702 -1.62368886 -0.8660254 -0.11290351
#> [19,]  0.34165840 -1.12379931 -0.8660254 -0.64932289
#> [20,]  2.74327390 -1.57369991 -0.8660254  0.42351587
#> [21,] -0.64900800  0.42585829  1.1547005 -0.63591240
#> attr(,"center")
#>  [1]   0.3809524   3.0952381   6.2857143 230.0571429   3.5638095   3.7619048
#>  [7] 152.2380952  17.7480952   0.4285714   3.2541905
#> attr(,"scale")
#>  [1]   0.4856209   1.6876233   1.6659863 110.2307043   0.4523639   0.7499055
#>  [7]  66.6218217   2.0004419   0.4948717   0.7456852
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 29
#> 
#> $lambda.min
#> [1] 0.5961666
#> 
#> $null.dev
#> [1] 24.92503
#> 
#> $Bias
#> [1] 1.580626
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#>  9.43942