Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 34.832226 32.924640 29.788024 26.184807 22.433815 19.324543 16.763085
#>   [8] 14.668459 12.970885 11.610273 10.534926  9.700405  9.068552  8.606637
#>  [15]  8.286617  8.079874  7.875362  7.777935  7.749169  7.779158  7.824908
#>  [22]  7.744975  7.655308  7.541456  7.467984  7.432173  7.427228  7.447361
#>  [29]  7.487660  7.543962  7.612750  7.691065  7.713087  7.907668  9.759946
#>  [36]  9.808339 10.119484  9.826911  9.848641 10.298582 10.035097 10.040931
#>  [43] 10.590749 10.365983 11.338755 12.310871 12.348302 13.874071 15.128398
#>  [50] 16.578913 16.575339 16.598571 16.582573 16.628030 16.628836 16.389809
#>  [57] 16.502176 16.466112 16.467047 16.467893 16.453261 15.992714 15.862147
#>  [64] 15.862646 16.007556 15.941602 15.941681 17.880220 17.881628 17.950176
#>  [71] 17.953061 18.669112 18.687625 18.685915 18.684255 18.682645 18.681083
#>  [78] 18.317586 18.315846 18.314170 18.312556 18.311001 18.309504 18.308062
#>  [85] 18.306674 18.305337 18.304050 18.302811 18.301618 18.300470 18.297783
#>  [92] 18.198281 18.196414 18.194649 18.192971 18.191276 18.183430 18.181870
#>  [99] 18.180365 18.178908
#> 
#> $cvse
#>   [1] 9.372661 9.282364 8.728144 7.848994 6.785563 5.940508 5.271631 4.741141
#>   [9] 4.316278 3.969880 3.680562 3.432409 3.214269 3.018877 2.841961 2.681347
#>  [17] 2.541738 2.426215 2.334201 2.332683 2.350697 2.354707 2.349283 2.312075
#>  [25] 2.281624 2.260831 2.249498 2.247270 2.253618 2.267848 2.289135 2.316576
#>  [33] 2.324105 2.304118 2.512949 2.491378 2.482591 2.519736 2.513185 2.576954
#>  [41] 2.603576 2.603709 2.778253 2.724203 3.018172 2.974665 2.973325 3.748608
#>  [49] 3.908850 3.944995 3.946613 3.948599 3.971325 3.987608 3.987776 4.008937
#>  [57] 3.967319 3.953351 3.953500 3.953639 3.956371 3.965148 3.982916 3.983048
#>  [65] 3.949401 3.951972 3.951986 5.417310 5.418479 5.434618 5.436789 5.999787
#>  [73] 6.038953 6.037634 6.036346 6.035090 6.033866 5.997943 5.996737 5.995565
#>  [81] 5.994427 5.993322 5.992250 5.991209 5.990200 5.989222 5.988274 5.987355
#>  [89] 5.986465 5.985603 5.983664 5.914139 5.912665 5.911273 5.909950 5.908613
#>  [97] 5.902293 5.901074 5.899903 5.898773
#> 
#> $fold
#>  [1]  9  9  2  1  7 10  3  6  1  8  5  5  2  3  4  6  7  4  8  1 10
#> 
#> $lambda
#>   [1] 5.076072032 4.733961766 4.414908587 4.117358524 3.839862340 3.581068470
#>   [7] 3.339716440 3.114630730 2.904715043 2.708946971 2.526373012 2.356103926
#>  [13] 2.197310407 2.049219040 1.911108535 1.782306216 1.662184742 1.550159054
#>  [19] 1.445683522 1.348249291 1.257381801 1.172638476 1.093606567 1.019901145
#>  [25] 0.951163221 0.887058004 0.827273263 0.771517814 0.719520095 0.671026849
#>  [31] 0.625801885 0.583624933 0.544290566 0.507607203 0.473396177 0.441490859
#>  [37] 0.411735852 0.383986234 0.358106847 0.333971644 0.311463073 0.290471504
#>  [43] 0.270894697 0.252637301 0.235610393 0.219731041 0.204921904 0.191110853
#>  [49] 0.178230622 0.166218474 0.155015905 0.144568352 0.134824929 0.125738180
#>  [55] 0.117263848 0.109360657 0.101990114 0.095116322 0.088705800 0.082727326
#>  [61] 0.077151781 0.071952010 0.067102685 0.062580189 0.058362493 0.054429057
#>  [67] 0.050760720 0.047339618 0.044149086 0.041173585 0.038398623 0.035810685
#>  [73] 0.033397164 0.031146307 0.029047150 0.027089470 0.025263730 0.023561039
#>  [79] 0.021973104 0.020492190 0.019111085 0.017823062 0.016621847 0.015501591
#>  [85] 0.014456835 0.013482493 0.012573818 0.011726385 0.010936066 0.010199011
#>  [91] 0.009511632 0.008870580 0.008272733 0.007715178 0.007195201 0.006710268
#>  [97] 0.006258019 0.005836249 0.005442906 0.005076072
#> 
#> $fit
#> $beta
#>               5.0761     4.7340    4.4149    4.1174    3.8399    3.5811
#> (Intercept) 20.57619 22.4449460 24.187754 25.813102 27.328907 28.742552
#> am           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000 -0.5977467 -1.155207 -1.675097 -2.159948 -2.612121
#>               3.3397    3.1146    2.9047    2.7089    2.5264    2.3561
#> (Intercept) 30.06092 31.290439 32.437090 33.506460 34.503759 35.433843
#> am           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -3.03382 -3.427097 -3.793869 -4.135922 -4.454921 -4.752421
#>                2.1973    2.0492    1.9111   1.7823    1.6622    1.5502
#> (Intercept) 36.301243 37.110182 37.864602 38.56818 39.061337 39.061337
#> am           0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> wt          -5.029871 -5.288621 -5.529932 -5.75498 -5.912724 -5.912724
#>                1.4457    1.3482      1.2574     1.1726    1.0936     1.0199
#> (Intercept) 39.061337 39.061337 38.58367703 37.1276419 35.769496 34.5028844
#> am           0.000000  0.000000  0.00000000  0.0000000  0.000000  0.0000000
#> carb         0.000000  0.000000  0.00000000  0.0000000  0.000000  0.0000000
#> cyl          0.000000  0.000000  0.00000000  0.0000000  0.000000  0.0000000
#> disp         0.000000  0.000000  0.00000000  0.0000000  0.000000  0.0000000
#> drat         0.000000  0.000000  0.00000000  0.0000000  0.000000  0.0000000
#> gear         0.000000  0.000000  0.00000000  0.0000000  0.000000  0.0000000
#> hp           0.000000  0.000000  0.00000000  0.0000000  0.000000  0.0000000
#> qsec         0.000000  0.000000  0.02475349  0.1002087  0.170591  0.2362298
#> vs           0.000000  0.000000  0.00000000  0.0000000  0.000000  0.0000000
#> wt          -5.912724 -5.912724 -5.90218217 -5.8700471 -5.840073 -5.8121181
#>                 0.9512     0.8871     0.8273     0.7715     0.7195     0.6710
#> (Intercept) 33.3216383 32.2200041 31.1926165 30.2344713 29.3409019 28.5075561
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.2974448  0.3545341  0.4077757  0.4574291  0.5037359  0.5469219
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.7860477 -5.7617344 -5.7390597 -5.7179133 -5.6981920 -5.6797998
#>                 0.6258     0.5836     0.5443     0.5076     0.4734     0.4415
#> (Intercept) 27.7303750 27.0055733 26.3296208 25.6992253 25.1113163 24.6561570
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.5871972  0.6247582  0.6597876  0.6924562  0.7229231  0.7465105
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.6626473 -5.6466507 -5.6317323 -5.6178193 -5.6048440 -5.5947986
#>                 0.4117     0.3840      0.3581      0.3340      0.3115
#> (Intercept) 24.6561260 24.6561247 34.79050042 34.06272522 33.38445549
#> am           0.0000000  0.0000000  0.00000000  0.00000000  0.00000000
#> carb         0.0000000  0.0000000  0.00000000  0.00000000  0.00000000
#> cyl          0.0000000  0.0000000  0.00000000  0.00000000  0.00000000
#> disp         0.0000000  0.0000000  0.00000000  0.00000000  0.00000000
#> drat         0.0000000  0.0000000  0.00000000  0.00000000  0.00000000
#> gear         0.0000000  0.0000000  0.69653910  0.83391103  0.96193989
#> hp           0.0000000  0.0000000 -0.03057478 -0.03101215 -0.03141941
#> qsec         0.7465121  0.7465122  0.00000000  0.00000000  0.00000000
#> vs           0.0000000  0.0000000  0.00000000  0.00000000  0.00000000
#> wt          -5.5947979 -5.5947978 -3.96133232 -3.86959920 -3.78412405
#>                  0.2905      0.2709      0.2526      0.2356      0.2197
#> (Intercept) 32.75470743 32.32166629 32.32030953 32.31948119 32.31897545
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         1.08081851  1.16255300  1.16279876  1.16294881  1.16304042
#> hp          -0.03179534 -0.03205676 -0.03206021 -0.03206232 -0.03206361
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -3.70487448 -3.65023182 -3.64992762 -3.64974190 -3.64962851
#>                  0.2049      0.1911      0.1782      0.1662      0.1550
#> (Intercept) 36.88992420 36.89035499 36.88999928 36.88932278 36.88858486
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat        -1.26527984 -1.26485675 -1.26443358 -1.26406567 -1.26377162
#> gear         1.50188843  1.50141926  1.50109332  1.50087685  1.50073978
#> hp          -0.03419626 -0.03419081 -0.03418714 -0.03418479 -0.03418336
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -3.94064795 -3.94097747 -3.94114146 -3.94120671 -3.94121735
#>                  0.1446      0.1348      0.1257      0.1173      0.1094
#> (Intercept) 36.88791468 18.52414227 22.25518813 22.25866483 22.26163521
#> am           0.00000000  1.36547483  1.54919941  1.54843764  1.54777443
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.01974578  0.01785167  0.01784940  0.01784781
#> drat        -1.26355048  0.00000000 -0.87814813 -0.87807263 -0.87802072
#> gear         1.50065768  1.51942588  1.60290995  1.60325518  1.60357467
#> hp          -0.03418254 -0.03089330 -0.03162373 -0.03163101 -0.03163758
#> qsec         0.00000000  0.65989536  0.64302595  0.64276997  0.64255126
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -3.94120109 -5.07798798 -5.10525843 -5.10480337 -5.10443587
#>                  0.1020      0.0951      0.0887      0.0827      0.0772
#> (Intercept) 22.26415182 22.26626462 22.26802060 22.26946351 13.61389417
#> am           1.54720137  1.54671011  1.54629247  1.54594060  1.80042302
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.75943918
#> disp         0.01784679  0.01784621  0.01784601  0.01784609  0.01735450
#> drat        -0.87798828 -0.87797168 -0.87796779 -0.87797388 -0.54842130
#> gear         1.60386767  1.60413412  1.60437451  1.60458975  1.98063456
#> hp          -0.03164347 -0.03164871 -0.03165334 -0.03165740 -0.03659017
#> qsec         0.64236597  0.64221042  0.64208118  0.64197501  0.78217084
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.82362014
#> wt          -5.10414392 -5.10391666 -5.10374433 -5.10361822 -5.31733691
#>                  0.0720      0.0671      0.0626      0.0584      0.0544
#> (Intercept)  7.54417581  7.54056214  7.53722623  7.53415452  7.53133354
#> am           1.78713547  1.78736552  1.78758540  1.78779495  1.78799409
#> carb        -0.55170376 -0.55177768 -0.55184306 -0.55190067 -0.55195121
#> cyl          1.03930925  1.03956050  1.03978513  1.03998508  1.04016220
#> disp         0.01302546  0.01302693  0.01302841  0.01302989  0.01303136
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         2.54445354  2.54468632  2.54489220  2.54507333  2.54523175
#> hp          -0.02834524 -0.02834568 -0.02834600 -0.02834622 -0.02834635
#> qsec         0.75065693  0.75074049  0.75082149  0.75089969  0.75097491
#> vs           1.03775880  1.03806435  1.03833735  1.03858017  1.03879507
#> wt          -4.65984638 -4.66002197 -4.66019645 -4.66036864 -4.66053751
#>                  0.0508      0.0473      0.0441      0.0412      0.0384
#> (Intercept)  7.52874991  7.52639044  7.52424218  7.52229246  7.52052893
#> am           1.78818281  1.78836117  1.78852929  1.78868733  1.78883550
#> carb        -0.55199534 -0.55203366 -0.55206674 -0.55209508 -0.55211918
#> cyl          1.04031823  1.04045487  1.04057370  1.04067622  1.04076385
#> disp         0.01303281  0.01303424  0.01303562  0.01303697  0.01303827
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         2.54536937  2.54548803  2.54558941  2.54567512  2.54574665
#> hp          -0.02834639 -0.02834636 -0.02834625 -0.02834610 -0.02834589
#> qsec         0.75104699  0.75111583  0.75118136  0.75124354  0.75130236
#> vs           1.03898419  1.03914958  1.03929320  1.03941688  1.03952236
#> wt          -4.66070219 -4.66086197 -4.66101625 -4.66116457 -4.66130655
#>                  0.0358      0.0334      0.0311      0.0290      0.0271
#> (Intercept)  7.51893959  7.51751282  7.51623743  7.51510263  7.51409808
#> am           1.78897404  1.78910323  1.78922338  1.78933480  1.78943784
#> carb        -0.55213948 -0.55215637 -0.55217024 -0.55218142 -0.55219024
#> cyl          1.04083792  1.04089971  1.04095038  1.04099103  1.04102271
#> disp         0.01303952  0.01304071  0.01304185  0.01304293  0.01304396
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         2.54580539  2.54585264  2.54588962  2.54591743  2.54593712
#> hp          -0.02834563 -0.02834534 -0.02834502 -0.02834467 -0.02834431
#> qsec         0.75135783  0.75140998  0.75145886  0.75150454  0.75154709
#> vs           1.03961128  1.03968518  1.03974550  1.03979360  1.03983075
#> wt          -4.66144192 -4.66157050 -4.66169217 -4.66180689 -4.66191467
#>                  0.0253      0.0236      0.0220      0.0205      0.0191
#> (Intercept)  7.51321389  7.51244061  7.51176927  7.51119134  7.51069877
#> am           1.78953285  1.78962019  1.78970024  1.78977336  1.78983991
#> carb        -0.55219696 -0.55220186 -0.55220518 -0.55220711 -0.55220787
#> cyl          1.04104636  1.04106287  1.04107307  1.04107772  1.04107752
#> disp         0.01304492  0.01304583  0.01304668  0.01304747  0.01304820
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         2.54594963  2.54595586  2.54595662  2.54595263  2.54594457
#> hp          -0.02834392 -0.02834353 -0.02834313 -0.02834272 -0.02834231
#> qsec         0.75158662  0.75162322  0.75165701  0.75168809  0.75171658
#> vs           1.03985810  1.03987676  1.03988774  1.03989196  1.03989029
#> wt          -4.66201558 -4.66210972 -4.66219723 -4.66227829 -4.66235310
#>                  0.0178      0.0166      0.0155      0.0145      0.0135
#> (Intercept)  7.51028393  7.50993966  7.50965922  7.50943629  7.50926498
#> am           1.78990028  1.78995482  1.79000390  1.79004785  1.79008702
#> carb        -0.55220762 -0.55220653 -0.55220473 -0.55220235 -0.55219950
#> cyl          1.04107312  1.04106510  1.04105400  1.04104032  1.04102450
#> disp         0.01304888  0.01304951  0.01305009  0.01305062  0.01305110
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         2.54593307  2.54591868  2.54590190  2.54588319  2.54586296
#> hp          -0.02834191 -0.02834151 -0.02834111 -0.02834072 -0.02834034
#> qsec         0.75174262  0.75176631  0.75178779  0.75180718  0.75182461
#> vs           1.03988352  1.03987237  1.03985750  1.03983951  1.03981895
#> wt          -4.66242188 -4.66248487 -4.66254232 -4.66259449 -4.66264165
#>                  0.0126      0.0117      0.0109      0.0102      0.0095
#> (Intercept)  6.60556128  6.60098209  6.59665250  6.59256140  6.58869803
#> am           1.77367235  1.77370746  1.77374335  1.77377976  1.77381644
#> carb        -0.58702300 -0.58722698 -0.58741875 -0.58759898 -0.58776835
#> cyl          1.06998893  1.07026604  1.07052695  1.07077245  1.07100329
#> disp         0.01295429  0.01295380  0.01295336  0.01295298  0.01295266
#> drat         0.13330423  0.13362292  0.13392225  0.13420328  0.13446702
#> gear         2.57698813  2.57734118  2.57767314  2.57798515  2.57827825
#> hp          -0.02780483 -0.02780358 -0.02780239 -0.02780125 -0.02780018
#> qsec         0.75349861  0.75351593  0.75353353  0.75355128  0.75356908
#> vs           1.06601645  1.06634894  1.06666178  1.06695596  1.06723241
#> wt          -4.62016437 -4.62001856 -4.61988494 -4.61976255 -4.61965049
#>                  0.0089      0.0083      0.0077      0.0072      0.0067
#> (Intercept)  6.58505203  6.58161337  6.57837234  6.57531958  6.57244605
#> am           1.77385319  1.77388980  1.77392611  1.77396196  1.77399722
#> carb        -0.58792745 -0.58807687 -0.58821716 -0.58834885 -0.58847241
#> cyl          1.07122018  1.07142383  1.07161490  1.07179404  1.07196188
#> disp         0.01295237  0.01295213  0.01295193  0.01295176  0.01295161
#> drat         0.13471444  0.13494645  0.13516390  0.13536764  0.13555843
#> gear         2.57855348  2.57881179  2.57905409  2.57928128  2.57949416
#> hp          -0.02779915 -0.02779818 -0.02779725 -0.02779637 -0.02779553
#> qsec         0.75358681  0.75360440  0.75362176  0.75363882  0.75365554
#> vs           1.06749204  1.06773572  1.06796428  1.06817852  1.06837919
#> wt          -4.61954793 -4.61945410 -4.61936828 -4.61928981 -4.61921806
#>                  0.0063      0.0058      0.0054      0.0051
#> (Intercept)  6.56974301  6.56720204  6.56481499  6.56257405
#> am           1.77403178  1.77406554  1.77409842  1.77413035
#> carb        -0.58858831 -0.58869700 -0.58879889 -0.58889437
#> cyl          1.07211902  1.07226603  1.07240347  1.07253188
#> disp         0.01295149  0.01295140  0.01295132  0.01295127
#> drat         0.13573702  0.13590413  0.13606042  0.13620652
#> gear         2.57969356  2.57988021  2.58005484  2.58021813
#> hp          -0.02779472 -0.02779396 -0.02779324 -0.02779255
#> qsec         0.75367185  0.75368772  0.75370311  0.75371800
#> vs           1.06856705  1.06874277  1.06890704  1.06906050
#> wt          -4.61915247 -4.61909251 -4.61903770 -4.61898760
#> 
#> $iter
#>   [1]   0   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   4   3   3   3   3   3   3   3   3   3   3   3   3   3   3   3
#>  [37]   1   1  52  14  14  13   9   1   1   1  45   1   1   1   1   1  71  53
#>  [55]   1   1   1   1   1   1 150  84   1   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1  51   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.076072032 4.733961766 4.414908587 4.117358524 3.839862340 3.581068470
#>   [7] 3.339716440 3.114630730 2.904715043 2.708946971 2.526373012 2.356103926
#>  [13] 2.197310407 2.049219040 1.911108535 1.782306216 1.662184742 1.550159054
#>  [19] 1.445683522 1.348249291 1.257381801 1.172638476 1.093606567 1.019901145
#>  [25] 0.951163221 0.887058004 0.827273263 0.771517814 0.719520095 0.671026849
#>  [31] 0.625801885 0.583624933 0.544290566 0.507607203 0.473396177 0.441490859
#>  [37] 0.411735852 0.383986234 0.358106847 0.333971644 0.311463073 0.290471504
#>  [43] 0.270894697 0.252637301 0.235610393 0.219731041 0.204921904 0.191110853
#>  [49] 0.178230622 0.166218474 0.155015905 0.144568352 0.134824929 0.125738180
#>  [55] 0.117263848 0.109360657 0.101990114 0.095116322 0.088705800 0.082727326
#>  [61] 0.077151781 0.071952010 0.067102685 0.062580189 0.058362493 0.054429057
#>  [67] 0.050760720 0.047339618 0.044149086 0.041173585 0.038398623 0.035810685
#>  [73] 0.033397164 0.031146307 0.029047150 0.027089470 0.025263730 0.023561039
#>  [79] 0.021973104 0.020492190 0.019111085 0.017823062 0.016621847 0.015501591
#>  [85] 0.014456835 0.013482493 0.012573818 0.011726385 0.010936066 0.010199011
#>  [91] 0.009511632 0.008870580 0.008272733 0.007715178 0.007195201 0.006710268
#>  [97] 0.006258019 0.005836249 0.005442906 0.005076072
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 38
#> 
#> $loss
#>   [1] 686.31810 582.44390 495.53769 423.15764 363.19559 313.83255 273.50001
#>   [8] 240.84646 214.70822 194.08419 178.11398 166.05879 157.28505 151.24999
#>  [15] 147.48935 145.60657 145.22144 145.22144 145.22144 145.22144 142.86150
#>  [22] 136.15876 130.57312 125.94366 122.13044 119.01275 116.48661 114.46247
#>  [29] 112.86328 111.62282 110.68423 109.99873 109.52454 109.22591 109.07230
#>  [36] 109.03617 109.03617 109.03617  90.82131  90.14460  89.72422  89.51555
#>  [43]  89.47356  89.47356  89.47355  89.47355  86.59077  86.59076  86.59076
#>  [50]  86.59076  86.59076  86.59076  77.13996  75.91121  75.91119  75.91118
#>  [57]  75.91117  75.91116  75.91116  75.91115  73.27517  71.80226  71.80225
#>  [64]  71.80224  71.80224  71.80223  71.80223  71.80223  71.80222  71.80222
#>  [71]  71.80222  71.80222  71.80222  71.80221  71.80221  71.80221  71.80221
#>  [78]  71.80221  71.80221  71.80221  71.80221  71.80221  71.80221  71.80221
#>  [85]  71.80221  71.80221  71.78292  71.78291  71.78290  71.78290  71.78289
#>  [92]  71.78288  71.78288  71.78287  71.78287  71.78286  71.78286  71.78286
#>  [99]  71.78285  71.78285
#> 
#> $linear.predictors
#>      5.0761   4.7340   4.4149   4.1174   3.8399   3.5811   3.3397   3.1146
#> 1  20.57619 20.72642 20.86653 20.99720 21.11906 21.23270 21.33869 21.43753
#> 2  20.57619 21.05817 21.50767 21.92688 22.31783 22.68243 23.02246 23.33957
#> 3  20.57619 20.52319 20.47376 20.42767 20.38468 20.34458 20.30719 20.27232
#> 4  20.57619 20.38870 20.21384 20.05077 19.89869 19.75686 19.62458 19.50122
#> 5  20.57619 20.31099 20.06366 19.83301 19.61789 19.41728 19.23019 19.05570
#> 6  20.57619 20.53813 20.50264 20.46954 20.43867 20.40989 20.38304 20.35800
#> 7  20.57619 20.56204 20.54885 20.53655 20.52507 20.51437 20.50439 20.49508
#> 8  20.57619 20.38870 20.21384 20.05077 19.89869 19.75686 19.62458 19.50122
#> 9  20.57619 20.01212 19.48606 18.99546 18.53792 18.11122 17.71328 17.34215
#> 10 20.57619 20.21535 19.87883 19.56499 19.27230 18.99934 18.74477 18.50737
#> 11 20.57619 19.30678 18.12292 17.01884 15.98918 15.02892 14.13337 13.29818
#> 12 20.57619 21.12990 21.64630 22.12789 22.57702 22.99589 23.38652 23.75082
#> 13 20.57619 21.47959 22.32209 23.10782 23.84059 24.52398 25.16130 25.75568
#> 14 20.57619 20.97150 21.34017 21.68399 22.00464 22.30367 22.58256 22.84264
#> 15 20.57619 20.39169 20.21962 20.05914 19.90949 19.76992 19.63975 19.51836
#> 16 20.57619 20.14960 19.75176 19.38073 19.03471 18.71201 18.41105 18.13039
#> 17 20.57619 20.14661 19.74598 19.37235 19.02391 18.69895 18.39589 18.11325
#> 18 20.57619 21.16577 21.71561 22.22839 22.70662 23.15261 23.56855 23.95645
#> 19 20.57619 21.54056 22.43992 23.27868 24.06091 24.79041 25.47075 26.10524
#> 20 20.57619 20.31099 20.06366 19.83301 19.61789 19.41728 19.23019 19.05570
#> 21 20.57619 20.78321 20.97628 21.15633 21.32425 21.48086 21.62690 21.76311
#>      2.9047   2.7089   2.5264   2.3561    2.1973    2.0492    1.9111    1.7823
#> 1  21.52972 21.61569 21.69586 21.77063 21.840365 21.905397 21.966047 22.022609
#> 2  23.63531 23.91112 24.16834 24.40823 24.631943 24.840582 25.035159 25.216623
#> 3  20.23980 20.20947 20.18119 20.15481 20.130209 20.107266 20.085870 20.065916
#> 4  19.38618 19.27889 19.17883 19.08551 18.998488 18.917326 18.841635 18.771045
#> 5  18.89298 18.74122 18.59969 18.46770 18.344605 18.229806 18.122744 18.022898
#> 6  20.33465 20.31287 20.29256 20.27362 20.255955 20.239482 20.224118 20.209790
#> 7  20.48640 20.47831 20.47076 20.46372 20.457150 20.451026 20.445315 20.439989
#> 8  19.38618 19.27889 19.17883 19.08551 18.998488 18.917326 18.841635 18.771045
#> 9  16.99604 16.67326 16.37223 16.09149 15.829669 15.585495 15.357778 15.145408
#> 10 18.28596 18.07947 17.88690 17.70731 17.539825 17.383626 17.237955 17.102101
#> 11 12.51928 11.79287 11.11542 10.48363  9.894422  9.344923  8.832458  8.354531
#> 12 24.09058 24.40743 24.70293 24.97852 25.235527 25.475216 25.698751 25.907220
#> 13 26.30999 26.82695 27.30906 27.75868 28.178002 28.569060 28.933761 29.273884
#> 14 23.08520 23.31141 23.52238 23.71912 23.902612 24.073732 24.233319 24.382151
#> 15 19.40515 19.29957 19.20110 19.10928 19.023637 18.943769 18.869285 18.799820
#> 16 17.86863 17.62452 17.39686 17.18455 16.986539 16.801878 16.629662 16.469053
#> 17 17.84966 17.60384 17.37459 17.16078 16.961390 16.775435 16.602012 16.440278
#> 18 24.31821 24.65559 24.97023 25.26366 25.537320 25.792534 26.030547 26.252519
#> 19 26.69697 27.24881 27.76346 28.24343 28.691048 29.108499 29.497815 29.860892
#> 20 18.89298 18.74122 18.59969 18.46770 18.344605 18.229806 18.122744 18.022898
#> 21 21.89013 22.00860 22.11908 22.22211 22.318202 22.407816 22.491390 22.569332
#>       1.6622    1.5502    1.4457    1.3482    1.2574    1.1726    1.0936
#> 1  22.062255 22.062255 22.062255 22.062255 22.036208 21.956808 21.882746
#> 2  25.343817 25.343817 25.343817 25.343817 25.351277 25.374016 25.395226
#> 3  20.051929 20.051929 20.051929 20.051929 20.089369 20.203497 20.309952
#> 4  18.721566 18.721566 18.721566 18.721566 18.701475 18.640231 18.583105
#> 5  17.952912 17.952912 17.952912 17.952912 17.904982 17.758879 17.622598
#> 6  20.199747 20.199747 20.199747 20.199747 20.250786 20.406365 20.551485
#> 7  20.436256 20.436256 20.436256 20.436256 20.558658 20.931772 21.279801
#> 8  18.721566 18.721566 18.721566 18.721566 18.748011 18.828624 18.903816
#> 9  14.996550 14.996550 14.996550 14.996550 14.992506 14.980181 14.968684
#> 10 17.006876 17.006876 17.006876 17.006876 17.004199 16.996039 16.988427
#> 11  8.019535  8.019535  8.019535  8.019535  8.042288  8.111646  8.176341
#> 12 26.053344 26.053344 26.053344 26.053344 26.080827 26.164601 26.242743
#> 13 29.512288 29.512288 29.512288 29.512288 29.510087 29.503380 29.497124
#> 14 24.486472 24.486472 24.486472 24.486472 24.530115 24.663151 24.787243
#> 15 18.751130 18.751130 18.751130 18.751130 18.737917 18.697640 18.660071
#> 16 16.356476 16.356476 16.356476 16.356476 16.300749 16.130876 15.972425
#> 17 16.326913 16.326913 16.326913 16.326913 16.311834 16.265868 16.222994
#> 18 26.408107 26.408107 26.408107 26.408107 26.366390 26.239226 26.120610
#> 19 30.115386 30.115386 30.115386 30.115386 30.072009 29.939787 29.816454
#> 20 17.952912 17.952912 17.952912 17.952912 17.874288 17.634620 17.411066
#> 21 22.623964 22.623964 22.623964 22.623964 22.636025 22.672792 22.707087
#>       1.0199    0.9512    0.8871   0.8273    0.7715    0.7195    0.6710
#> 1  21.813676 21.749261 21.689187 21.63316 21.580913 21.532186 21.486742
#> 2  25.415007 25.433455 25.450659 25.46670 25.481667 25.495622 25.508637
#> 3  20.409232 20.501821 20.588170 20.66870 20.743801 20.813841 20.879161
#> 4  18.529829 18.480144 18.433807 18.39059 18.350292 18.312707 18.277655
#> 5  17.495503 17.376973 17.266432 17.16334 17.067197 16.977534 16.893913
#> 6  20.686824 20.813041 20.930752 21.04053 21.142909 21.238388 21.327432
#> 7  21.604375 21.907073 22.189371 22.45264 22.698170 22.927150 23.140698
#> 8  18.973941 19.039340 19.100331 19.15721 19.210259 19.259731 19.305868
#> 9  14.957962 14.947963 14.938638 14.92994 14.921830 14.914266 14.907212
#> 10 16.981328 16.974708 16.968534 16.96278 16.957406 16.952398 16.947728
#> 11  8.236676  8.292945  8.345421  8.39436  8.440001  8.482566  8.522263
#> 12 26.315619 26.383583 26.446966 26.50608 26.561206 26.612618 26.660566
#> 13 29.491290 29.485848 29.480774 29.47604 29.471628 29.467511 29.463673
#> 14 24.902972 25.010900 25.111555 25.20543 25.292971 25.374615 25.450756
#> 15 18.625034 18.592359 18.561886 18.53347 18.506962 18.482244 18.459192
#> 16 15.824652 15.686839 15.558314 15.43845 15.326666 15.222416 15.125191
#> 17 16.183008 16.145718 16.110941 16.07851 16.048260 16.020052 15.993744
#> 18 26.009989 25.906824 25.810611 25.72088 25.637202 25.559161 25.486380
#> 19 29.701433 29.594165 29.494126 29.40083 29.313820 29.232675 29.156999
#> 20 17.202578 17.008141 16.826809 16.65770 16.499985 16.352901 16.215730
#> 21 22.739070 22.768898 22.796716 22.82266 22.846853 22.869417 22.890460
#>       0.6258    0.5836    0.5443    0.5076    0.4734    0.4415    0.4117
#> 1  21.444361 21.404837 21.367976 21.333600 21.301540 21.276720 21.276718
#> 2  25.520774 25.532093 25.542650 25.552495 25.561676 25.568785 25.568785
#> 3  20.940079 20.996890 21.049873 21.099285 21.145367 21.181044 21.181046
#> 4  18.244966 18.214479 18.186047 18.159532 18.134803 18.115659 18.115657
#> 5  16.815929 16.743200 16.675373 16.612117 16.553124 16.507452 16.507449
#> 6  21.410475 21.487921 21.560148 21.627506 21.690325 21.738959 21.738963
#> 7  23.339853 23.525586 23.698801 23.860342 24.010996 24.127632 24.127640
#> 8  19.348896 19.389025 19.426448 19.461350 19.493899 19.519098 19.519100
#> 9  14.900633 14.894497 14.888775 14.883439 14.878462 14.874609 14.874609
#> 10 16.943372 16.939310 16.935522 16.931989 16.928694 16.926143 16.926143
#> 11  8.559283  8.593809  8.626008  8.656037  8.684042  8.705723  8.705725
#> 12 26.705281 26.746984 26.785875 26.822146 26.855971 26.882159 26.882161
#> 13 29.460093 29.456754 29.453640 29.450736 29.448028 29.445932 29.445931
#> 14 25.521766 25.587991 25.649751 25.707350 25.761066 25.802653 25.802656
#> 15 18.437694 18.417645 18.398947 18.381509 18.365246 18.352655 18.352655
#> 16 15.034519 14.949958 14.871096 14.797550 14.728960 14.675857 14.675854
#> 17 15.969209 15.946328 15.924989 15.905089 15.886529 15.872160 15.872159
#> 18 25.418504 25.355202 25.296167 25.241111 25.189765 25.150013 25.150011
#> 19 29.086423 29.020604 28.959221 28.901975 28.848587 28.807254 28.807251
#> 20 16.087804 15.968500 15.857236 15.753471 15.656700 15.581779 15.581774
#> 21 22.910084 22.928387 22.945455 22.961373 22.976219 22.987712 22.987713
#>       0.3840    0.3581    0.3340    0.3115   0.2905   0.2709   0.2526   0.2356
#> 1  21.276718 22.824600 22.861935 22.896723 22.92898 22.95122 22.95134 22.95141
#> 2  25.568785 25.542911 25.536769 25.531042 25.52571 25.52206 25.52207 25.52208
#> 3  21.181046 20.781208 20.712360 20.648181 20.58850 20.54759 20.54756 20.54755
#> 4  18.115657 17.902548 17.825911 17.754492 17.68821 17.64259 17.64242 17.64231
#> 5  16.507449 15.247340 15.152013 15.063197 14.98090 14.92409 14.92371 14.92348
#> 6  21.738963 23.044370 23.131595 23.212856 23.28812 23.34012 23.34050 23.34074
#> 7  24.127640 22.193856 22.262978 22.327380 22.38707 22.42826 22.42851 22.42867
#> 8  19.519100 20.188975 20.272454 20.350241 20.42239 20.47210 20.47235 20.47250
#> 9  14.874609 15.254034 15.233003 15.213397 15.19516 15.18267 15.18266 15.18266
#> 10 16.926143 16.600887 16.548666 16.499999 16.45482 16.42374 16.42364 16.42357
#> 11  8.705725  9.815293  9.891572  9.962645 10.02853 10.07397 10.07424 10.07441
#> 12 26.882161 26.843790 26.838449 26.833461 26.82877 26.82562 26.82569 26.82573
#> 13 29.445931 29.589216 29.536335 29.487045 29.44125 29.40980 29.40974 29.40970
#> 14 25.802656 24.149680 24.017718 23.894727 23.78050 23.70200 23.70179 23.70167
#> 15 18.352655 18.686724 18.620563 18.558898 18.50162 18.46227 18.46217 18.46212
#> 16 14.675853 14.177780 14.107221 14.041484 13.98059 13.93853 13.93823 13.93805
#> 17 15.872159 16.298208 16.258723 16.221922 16.18774 16.16425 16.16420 16.16416
#> 18 25.150010 27.013640 27.129232 27.236963 27.33699 27.40577 27.40598 27.40611
#> 19 28.807251 28.824750 28.873204 28.918382 28.96045 28.98922 28.98916 28.98912
#> 20 15.581774 13.888688 14.028741 14.159330 14.28096 14.36409 14.36389 14.36377
#> 21 22.987713 23.231502 23.260559 23.287635 23.31274 23.33005 23.33014 23.33020
#>      0.2197    0.2049    0.1911    0.1782    0.1662    0.1550    0.1446
#> 1  22.95146 22.871935 22.871792 22.871714 22.871678 22.871665 22.871666
#> 2  25.52208 25.703595 25.703521 25.703451 25.703392 25.703347 25.703313
#> 3  20.54754 21.067755 21.067622 21.067468 21.067325 21.067204 21.067111
#> 4  17.64224 17.869783 17.869960 17.870037 17.870057 17.870048 17.870027
#> 5  14.92334 14.887843 14.888385 14.888723 14.888921 14.889029 14.889080
#> 6  23.34088 23.537760 23.537162 23.536769 23.536522 23.536375 23.536295
#> 7  22.42876 22.275895 22.275588 22.275419 22.275337 22.275306 22.275302
#> 8  20.47259 20.175611 20.175362 20.175248 20.175213 20.175219 20.175243
#> 9  15.18266 15.317416 15.317379 15.317337 15.317298 15.317266 15.317242
#> 10 16.42353 16.657236 16.657311 16.657325 16.657309 16.657280 16.657250
#> 11 10.07451  9.989684  9.989335  9.989132  9.989024  9.988974  9.988958
#> 12 26.82576 26.808757 26.808673 26.808621 26.808591 26.808576 26.808569
#> 13 29.40968 28.517296 28.517688 28.518040 28.518329 28.518549 28.518710
#> 14 23.70159 23.683319 23.683625 23.683809 23.683911 23.683963 23.683984
#> 15 18.46208 18.744393 18.744435 18.744421 18.744383 18.744338 18.744296
#> 16 13.93794 13.165923 13.166595 13.167109 13.167481 13.167739 13.167909
#> 17 16.16414 16.362390 16.362404 16.362385 16.362353 16.362319 16.362289
#> 18 27.40618 27.249330 27.249080 27.248952 27.248898 27.248885 27.248893
#> 19 28.98910 29.802883 29.802681 29.802457 29.802253 29.802083 29.801953
#> 20 14.36369 14.396414 14.396648 14.396804 14.396902 14.396961 14.396995
#> 21 23.33024 23.014784 23.014755 23.014779 23.014824 23.014872 23.014917
#>      0.1348   0.1257   0.1173   0.1094   0.1020   0.0951   0.0887   0.0827
#> 1  22.36059 22.43559 22.43577 22.43591 22.43603 22.43612 22.43619 22.43624
#> 2  25.72651 25.94464 25.94440 25.94418 25.94400 25.94383 25.94369 25.94356
#> 3  21.28120 21.57336 21.57321 21.57309 21.57301 21.57296 21.57293 21.57291
#> 4  18.54771 18.57241 18.57228 18.57219 18.57214 18.57212 18.57211 18.57213
#> 5  14.94637 14.88361 14.88333 14.88309 14.88289 14.88271 14.88256 14.88243
#> 6  22.58229 22.65937 22.66006 22.66066 22.66117 22.66162 22.66200 22.66232
#> 7  23.56313 23.37748 23.37719 23.37695 23.37673 23.37655 23.37640 23.37627
#> 8  19.11511 18.91781 18.91842 18.91892 18.91934 18.91969 18.91997 18.92019
#> 9  13.78228 14.00947 14.00968 14.00984 14.00995 14.01003 14.01008 14.01011
#> 10 15.64078 15.87387 15.87387 15.87385 15.87383 15.87381 15.87377 15.87374
#> 11 11.27478 11.19307 11.19303 11.19301 11.19300 11.19302 11.19304 11.19307
#> 12 27.15895 27.23909 27.23885 27.23864 27.23845 27.23829 27.23815 27.23802
#> 13 29.87594 29.25754 29.25745 29.25738 29.25731 29.25725 29.25720 29.25715
#> 14 23.14450 23.17374 23.17355 23.17338 23.17321 23.17305 23.17291 23.17278
#> 15 18.42444 18.56889 18.56899 18.56909 18.56919 18.56928 18.56937 18.56946
#> 16 13.09410 12.59353 12.59355 12.59355 12.59353 12.59349 12.59345 12.59340
#> 17 17.30076 17.29961 17.29956 17.29955 17.29956 17.29960 17.29966 17.29973
#> 18 27.20423 27.01182 27.01236 27.01284 27.01326 27.01363 27.01395 27.01423
#> 19 29.34286 29.77541 29.77546 29.77553 29.77561 29.77569 29.77578 29.77587
#> 20 14.58703 14.65210 14.65158 14.65110 14.65068 14.65031 14.64998 14.64969
#> 21 23.14644 23.08757 23.08741 23.08725 23.08710 23.08696 23.08682 23.08669
#>      0.0772   0.0720   0.0671   0.0626   0.0584   0.0544   0.0508   0.0473
#> 1  22.53165 21.88339 21.88325 21.88312 21.88300 21.88289 21.88279 21.88269
#> 2  25.77819 26.08194 26.08199 26.08204 26.08208 26.08212 26.08216 26.08219
#> 3  21.80962 21.75340 21.75362 21.75381 21.75400 21.75416 21.75431 21.75445
#> 4  18.76903 18.86366 18.86388 18.86408 18.86427 18.86445 18.86461 18.86477
#> 5  14.56060 14.28453 14.28445 14.28438 14.28432 14.28428 14.28425 14.28422
#> 6  22.33256 22.11524 22.11502 22.11481 22.11460 22.11441 22.11423 22.11406
#> 7  23.37755 23.46629 23.46630 23.46631 23.46633 23.46635 23.46637 23.46640
#> 8  19.66629 19.64294 19.64294 19.64292 19.64290 19.64286 19.64281 19.64275
#> 9  14.11601 14.42303 14.42297 14.42290 14.42283 14.42275 14.42268 14.42260
#> 10 16.08034 16.15751 16.15753 16.15753 16.15754 16.15753 16.15753 16.15751
#> 11 10.86219 10.65506 10.65504 10.65503 10.65503 10.65503 10.65504 10.65505
#> 12 27.44225 27.67036 27.67048 27.67058 27.67067 27.67076 27.67083 27.67090
#> 13 29.80387 29.48930 29.48937 29.48943 29.48949 29.48955 29.48960 29.48966
#> 14 22.46705 22.16982 22.16952 22.16924 22.16899 22.16877 22.16858 22.16841
#> 15 18.95753 19.07635 19.07652 19.07667 19.07681 19.07692 19.07703 19.07712
#> 16 12.32986 12.57334 12.57315 12.57299 12.57284 12.57270 12.57258 12.57247
#> 17 17.37154 17.51996 17.52017 17.52036 17.52054 17.52071 17.52087 17.52102
#> 18 26.36693 26.65885 26.65865 26.65846 26.65828 26.65812 26.65797 26.65783
#> 19 29.80060 29.81663 29.81681 29.81697 29.81712 29.81726 29.81738 29.81748
#> 20 14.85439 14.70337 14.70346 14.70354 14.70362 14.70370 14.70378 14.70385
#> 21 22.82197 23.09501 23.09492 23.09483 23.09475 23.09468 23.09461 23.09455
#>      0.0441   0.0412   0.0384   0.0358   0.0334   0.0311   0.0290   0.0271
#> 1  21.88260 21.88252 21.88244 21.88237 21.88231 21.88225 21.88220 21.88215
#> 2  26.08222 26.08225 26.08228 26.08231 26.08233 26.08235 26.08237 26.08239
#> 3  21.75458 21.75469 21.75480 21.75489 21.75497 21.75505 21.75511 21.75517
#> 4  18.86491 18.86504 18.86516 18.86527 18.86537 18.86547 18.86555 18.86563
#> 5  14.28421 14.28420 14.28420 14.28421 14.28422 14.28423 14.28425 14.28427
#> 6  22.11390 22.11375 22.11360 22.11347 22.11335 22.11323 22.11312 22.11302
#> 7  23.46643 23.46646 23.46649 23.46653 23.46656 23.46660 23.46663 23.46666
#> 8  19.64268 19.64261 19.64253 19.64246 19.64237 19.64229 19.64221 19.64212
#> 9  14.42252 14.42244 14.42236 14.42228 14.42221 14.42213 14.42206 14.42200
#> 10 16.15750 16.15748 16.15747 16.15745 16.15742 16.15740 16.15738 16.15736
#> 11 10.65507 10.65509 10.65512 10.65514 10.65517 10.65520 10.65523 10.65526
#> 12 27.67097 27.67102 27.67108 27.67112 27.67116 27.67120 27.67123 27.67126
#> 13 29.48971 29.48976 29.48981 29.48986 29.48990 29.48994 29.48997 29.49001
#> 14 22.16826 22.16813 22.16801 22.16791 22.16783 22.16776 22.16771 22.16766
#> 15 19.07720 19.07727 19.07733 19.07739 19.07743 19.07746 19.07749 19.07752
#> 16 12.57237 12.57228 12.57221 12.57214 12.57208 12.57203 12.57199 12.57195
#> 17 17.52116 17.52128 17.52140 17.52151 17.52161 17.52170 17.52178 17.52186
#> 18 26.65771 26.65760 26.65750 26.65740 26.65732 26.65725 26.65718 26.65713
#> 19 29.81758 29.81767 29.81774 29.81781 29.81787 29.81792 29.81797 29.81801
#> 20 14.70392 14.70399 14.70405 14.70411 14.70417 14.70422 14.70427 14.70432
#> 21 23.09450 23.09445 23.09441 23.09437 23.09434 23.09431 23.09429 23.09427
#>      0.0253   0.0236   0.0220   0.0205   0.0191   0.0178   0.0166   0.0155
#> 1  21.88210 21.88206 21.88202 21.88199 21.88196 21.88193 21.88191 21.88189
#> 2  26.08241 26.08243 26.08244 26.08245 26.08247 26.08248 26.08249 26.08250
#> 3  21.75522 21.75527 21.75531 21.75534 21.75537 21.75539 21.75541 21.75542
#> 4  18.86570 18.86576 18.86581 18.86586 18.86591 18.86595 18.86598 18.86601
#> 5  14.28429 14.28432 14.28434 14.28437 14.28440 14.28443 14.28445 14.28448
#> 6  22.11293 22.11285 22.11277 22.11270 22.11263 22.11258 22.11252 22.11248
#> 7  23.46670 23.46673 23.46676 23.46680 23.46683 23.46686 23.46688 23.46691
#> 8  19.64204 19.64196 19.64188 19.64180 19.64172 19.64165 19.64158 19.64151
#> 9  14.42193 14.42187 14.42181 14.42175 14.42170 14.42165 14.42160 14.42156
#> 10 16.15733 16.15731 16.15729 16.15726 16.15724 16.15722 16.15720 16.15718
#> 11 10.65529 10.65532 10.65535 10.65538 10.65541 10.65544 10.65546 10.65549
#> 12 27.67129 27.67131 27.67133 27.67134 27.67136 27.67137 27.67138 27.67138
#> 13 29.49004 29.49007 29.49010 29.49012 29.49014 29.49016 29.49018 29.49019
#> 14 22.16762 22.16760 22.16758 22.16756 22.16756 22.16756 22.16756 22.16757
#> 15 19.07753 19.07754 19.07755 19.07756 19.07756 19.07756 19.07755 19.07754
#> 16 12.57192 12.57189 12.57187 12.57185 12.57184 12.57183 12.57183 12.57183
#> 17 17.52192 17.52199 17.52204 17.52209 17.52213 17.52217 17.52221 17.52224
#> 18 26.65708 26.65703 26.65700 26.65696 26.65694 26.65691 26.65689 26.65688
#> 19 29.81804 29.81806 29.81809 29.81810 29.81812 29.81813 29.81814 29.81814
#> 20 14.70437 14.70441 14.70445 14.70448 14.70451 14.70454 14.70457 14.70460
#> 21 23.09425 23.09424 23.09422 23.09422 23.09421 23.09420 23.09420 23.09420
#>      0.0145   0.0135   0.0126   0.0117   0.0109   0.0102   0.0095   0.0089
#> 1  21.88187 21.88185 21.83464 21.83437 21.83412 21.83388 21.83365 21.83344
#> 2  26.08251 26.08252 26.07640 26.07645 26.07651 26.07656 26.07660 26.07665
#> 3  21.75544 21.75544 21.72389 21.72396 21.72402 21.72409 21.72415 21.72421
#> 4  18.86603 18.86606 18.87118 18.87131 18.87143 18.87155 18.87167 18.87178
#> 5  14.28451 14.28454 14.26904 14.26887 14.26871 14.26856 14.26842 14.26829
#> 6  22.11243 22.11240 22.08548 22.08533 22.08520 22.08507 22.08494 22.08482
#> 7  23.46694 23.46696 23.49210 23.49212 23.49214 23.49216 23.49218 23.49220
#> 8  19.64145 19.64139 19.67283 19.67299 19.67314 19.67328 19.67340 19.67352
#> 9  14.42152 14.42148 14.41934 14.41939 14.41943 14.41947 14.41950 14.41953
#> 10 16.15716 16.15714 16.14090 16.14090 16.14090 16.14090 16.14090 16.14090
#> 11 10.65552 10.65554 10.64540 10.64532 10.64524 10.64517 10.64510 10.64504
#> 12 27.67139 27.67139 27.68066 27.68076 27.68087 27.68096 27.68105 27.68114
#> 13 29.49021 29.49022 29.54432 29.54438 29.54443 29.54448 29.54453 29.54458
#> 14 22.16758 22.16759 22.13624 22.13591 22.13559 22.13529 22.13501 22.13475
#> 15 19.07754 19.07753 19.07494 19.07507 19.07519 19.07531 19.07542 19.07552
#> 16 12.57183 12.57183 12.63737 12.63740 12.63743 12.63745 12.63748 12.63750
#> 17 17.52226 17.52228 17.53146 17.53161 17.53175 17.53188 17.53201 17.53213
#> 18 26.65687 26.65686 26.68506 26.68505 26.68504 26.68502 26.68501 26.68500
#> 19 29.81814 29.81814 29.77248 29.77255 29.77261 29.77266 29.77272 29.77277
#> 20 14.70462 14.70464 14.69151 14.69149 14.69148 14.69146 14.69145 14.69143
#> 21 23.09420 23.09420 23.11475 23.11477 23.11478 23.11479 23.11480 23.11481
#>      0.0083   0.0077   0.0072   0.0067   0.0063   0.0058   0.0054   0.0051
#> 1  21.83324 21.83304 21.83286 21.83269 21.83254 21.83238 21.83224 21.83211
#> 2  26.07669 26.07673 26.07677 26.07681 26.07684 26.07688 26.07691 26.07694
#> 3  21.72426 21.72431 21.72436 21.72441 21.72446 21.72450 21.72454 21.72458
#> 4  18.87188 18.87198 18.87207 18.87216 18.87224 18.87232 18.87240 18.87247
#> 5  14.26817 14.26805 14.26795 14.26785 14.26776 14.26767 14.26759 14.26752
#> 6  22.08471 22.08460 22.08450 22.08440 22.08431 22.08422 22.08413 22.08405
#> 7  23.49222 23.49224 23.49226 23.49228 23.49230 23.49232 23.49234 23.49235
#> 8  19.67362 19.67372 19.67381 19.67389 19.67397 19.67403 19.67410 19.67415
#> 9  14.41956 14.41958 14.41960 14.41962 14.41963 14.41965 14.41966 14.41967
#> 10 16.14089 16.14089 16.14089 16.14088 16.14088 16.14088 16.14087 16.14087
#> 11 10.64498 10.64493 10.64488 10.64483 10.64479 10.64475 10.64471 10.64468
#> 12 27.68122 27.68129 27.68137 27.68143 27.68150 27.68156 27.68161 27.68167
#> 13 29.54462 29.54466 29.54470 29.54474 29.54477 29.54481 29.54484 29.54487
#> 14 22.13450 22.13427 22.13405 22.13385 22.13365 22.13347 22.13331 22.13315
#> 15 19.07562 19.07571 19.07580 19.07588 19.07595 19.07602 19.07609 19.07615
#> 16 12.63752 12.63753 12.63755 12.63756 12.63757 12.63759 12.63760 12.63761
#> 17 17.53225 17.53236 17.53246 17.53256 17.53265 17.53274 17.53283 17.53291
#> 18 26.68499 26.68497 26.68496 26.68495 26.68494 26.68492 26.68491 26.68490
#> 19 29.77282 29.77287 29.77291 29.77296 29.77300 29.77304 29.77307 29.77311
#> 20 14.69142 14.69142 14.69141 14.69140 14.69140 14.69140 14.69139 14.69139
#> 21 23.11482 23.11483 23.11484 23.11485 23.11485 23.11486 23.11486 23.11487
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 22.8 21.4 18.7 14.3 24.4 22.8 17.8 16.4 17.3 10.4 32.4 30.4 21.5 15.2
#> [16] 13.3 19.2 26.0 30.4 15.0 21.4
#> 
#> $X
#>               am       carb       cyl       disp       drat       gear
#>  [1,]  1.2747549  0.8419387  0.000000 -0.5394659  0.5221527  0.4677072
#>  [2,]  1.2747549 -1.0524234 -1.080123 -0.9821961  0.4240739  0.4677072
#>  [3,] -0.7844645 -1.0524234  0.000000  0.2949102 -1.0863391 -0.9354143
#>  [4,] -0.7844645 -0.4209693  1.080123  1.1633424 -0.9490289 -0.9354143
#>  [5,] -0.7844645  0.8419387  1.080123  1.1633424 -0.8313343 -0.9354143
#>  [6,] -0.7844645 -0.4209693 -1.080123 -0.6527026  0.1102219  0.4677072
#>  [7,] -0.7844645 -0.4209693 -1.080123 -0.7029355  0.5613842  0.4677072
#>  [8,] -0.7844645  0.8419387  0.000000 -0.4747592  0.5613842  0.4677072
#>  [9,] -0.7844645  0.2104847  1.080123  0.4464601 -1.1059549 -0.9354143
#> [10,] -0.7844645  0.2104847  1.080123  0.4464601 -1.1059549 -0.9354143
#> [11,] -0.7844645  0.8419387  1.080123  2.1169151 -1.3805755 -0.9354143
#> [12,]  1.2747549 -1.0524234 -1.080123 -1.2316575  0.8752363  0.4677072
#> [13,]  1.2747549 -0.4209693 -1.080123 -1.2571996  2.5425754  0.4677072
#> [14,] -0.7844645 -1.0524234 -1.080123 -0.8791762  0.1298376 -0.9354143
#> [15,] -0.7844645 -0.4209693  1.080123  0.6865561 -0.9490289 -0.9354143
#> [16,] -0.7844645  0.8419387  1.080123  1.0782020  0.1886849 -0.9354143
#> [17,] -0.7844645 -0.4209693  1.080123  1.5039041 -1.0863391 -0.9354143
#> [18,]  1.2747549 -0.4209693 -1.080123 -0.8774733  1.5617877  1.8708287
#> [19,]  1.2747549 -0.4209693 -1.080123 -1.0920272  0.2671479  1.8708287
#> [20,]  1.2747549  3.3677548  1.080123  0.6610140 -0.1840145  1.8708287
#> [21,]  1.2747549 -0.4209693 -1.080123 -0.8715135  0.9340835  0.4677072
#>                hp         qsec         vs          wt
#>  [1,] -0.48043955 -0.526297220 -0.9534626 -0.29275879
#>  [2,] -0.72511419  0.358995686  1.0488088 -0.93923542
#>  [3,] -0.48043955  0.821129719  1.0488088  0.10328095
#>  [4,]  0.45508112 -0.526297220 -0.9534626  0.36536607
#>  [5,]  1.46256492 -1.183307050 -0.9534626  0.51679303
#>  [6,] -1.17128559  1.132930994  1.0488088  0.07416038
#>  [7,] -0.69632894  2.747616169  1.0488088  0.02756747
#>  [8,] -0.29333542  0.520464203  1.0488088  0.36536607
#>  [9,]  0.52704425 -0.314717784 -0.9534626  1.09920441
#> [10,]  0.52704425 -0.203360185 -0.9534626  0.70316467
#> [11,]  0.88685989  0.008219251 -0.9534626  2.47369527
#> [12,] -1.11371509  0.837833358  1.0488088 -1.07901415
#> [13,] -1.31521185  0.308884767  1.0488088 -1.76043547
#> [14,] -0.66754369  1.138498874  1.0488088 -0.77033612
#> [15,]  0.09526548 -0.370396583 -0.9534626  0.35954196
#> [16,]  1.46256492 -1.422725886 -0.9534626  0.83129518
#> [17,]  0.45508112 -0.509593581 -0.9534626  0.83711929
#> [18,] -0.75389944 -0.704469378 -0.9534626 -1.14890352
#> [19,] -0.43726168 -0.593111779  1.0488088 -1.87924739
#> [20,]  2.75790123 -1.873724159 -0.9534626  0.51679303
#> [21,] -0.49483218  0.353427806  1.0488088 -0.40341695
#> attr(,"center")
#>  [1]   0.3809524   2.6666667   6.0000000 223.3619048   3.6338095   3.6666667
#>  [7] 143.3809524  17.9652381   0.4761905   3.1263333
#> attr(,"scale")
#>  [1]   0.4856209   1.5836466   1.8516402 117.4530321   0.5097943   0.7126966
#>  [7]  69.4800254   1.7960157   0.4994328   0.8584997
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 27
#> 
#> $lambda.min
#> [1] 0.8272733
#> 
#> $null.dev
#> [1] 32.68181
#> 
#> $Bias
#> [1] 3.480793
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 10.31488