Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 32.534464 31.485557 28.057389 24.473194 21.584265 19.284649 17.483229
#>   [8] 16.101725 15.072956 14.339332 13.851550 13.567458 13.451081 13.471764
#>  [15] 13.603439 13.741039 13.594945 13.313244 12.850627 12.498803 12.139027
#>  [22] 11.843783 11.577934 11.223645 10.924585 10.672920 10.461947 10.285826
#>  [29] 10.118045  9.916531 10.649122 10.528874 10.639126 10.936168 11.084406
#>  [36] 11.266785 11.220641 11.178401 10.758663 13.158838 14.574539 14.613406
#>  [43] 16.560957 17.752046 18.946940 18.487648 26.748971 28.918706 31.239448
#>  [50] 30.826492 30.780259 30.773696 31.965971 31.968005 32.053114 32.158395
#>  [57] 32.160088 31.818528 31.819972 31.573312 30.139347 30.139822 30.435848
#>  [64] 30.436739 30.965515 32.295926 32.648290 32.648559 32.714533 32.996555
#>  [71] 32.988807 32.988603 33.246461 33.243910 34.712198 34.712514 34.712850
#>  [78] 34.713206 34.713582 34.713976 34.713431 32.616890 32.616251 32.629865
#>  [85] 32.629432 32.628992 32.627847 32.681369 32.681601 32.681793 32.681950
#>  [92] 32.682072 32.682162 32.682221 32.682251 32.682253 32.682230 32.682183
#>  [99] 32.682113 32.682022
#> 
#> $cvse
#>   [1] 11.269343 11.261223 10.318628  9.157953  8.228379  7.473785  6.850071
#>   [8]  6.324709  5.875468  5.488648  5.157169  4.878697  4.653871  4.484658
#>  [15]  4.372834  4.280742  4.150598  4.077727  4.031400  3.984479  3.877329
#>  [22]  3.780017  3.685701  3.568887  3.462753  3.366501  3.279434  3.200936
#>  [29]  3.121274  3.024979  2.938293  2.891558  2.829944  2.815542  2.846833
#>  [36]  2.885586  2.879078  2.873241  2.656930  3.575399  3.587966  3.582826
#>  [43]  3.850159  4.106160  4.546880  4.576846  7.829679  9.391612 10.284189
#>  [50] 10.312689 10.318539 10.320220 10.439337 10.440121 10.437428 10.425264
#>  [57] 10.425862 10.451069 10.451557 10.463770  9.935113  9.935283  9.904144
#>  [64]  9.904251  9.874673 10.886272 10.985589 10.985050 10.976779 10.960661
#>  [71] 10.962702 10.962259 11.071428 11.070453 11.164209 11.163807 11.163427
#>  [78] 11.163069 11.162732 11.162416 11.161949 10.444912 10.443896 10.441147
#>  [85] 10.440166 10.439217 10.437654 10.430747 10.430398 10.430039 10.429669
#>  [92] 10.429290 10.428902 10.428508 10.428106 10.427699 10.427286 10.426870
#>  [99] 10.426449 10.426026
#> 
#> $fold
#>  [1]  2  9  7  9  6 10  2  8  5  6  5  1  3  1  4  3 10  4  1  8  7
#> 
#> $lambda
#>   [1] 4.744484503 4.424722127 4.126510664 3.848397656 3.589028535 3.347140023
#>   [7] 3.121553988 2.911171697 2.714968468 2.531988680 2.361341117 2.202194629
#>  [13] 2.053774081 1.915356582 1.786267959 1.665879477 1.553604776 1.448897014
#>  [19] 1.351246204 1.260176732 1.175245038 1.096037456 1.022168200 0.953277484
#>  [25] 0.889029772 0.829112141 0.773232758 0.721119458 0.672518420 0.627192929
#>  [31] 0.584922225 0.545500425 0.508735522 0.474448450 0.442472213 0.412651066
#>  [37] 0.384839766 0.358902853 0.334714002 0.312155399 0.291117170 0.271496847
#>  [43] 0.253198868 0.236134112 0.220219463 0.205377408 0.191535658 0.178626796
#>  [49] 0.166587948 0.155360478 0.144889701 0.135124620 0.126017673 0.117524504
#>  [55] 0.109603746 0.102216820 0.095327748 0.088902977 0.082911214 0.077323276
#>  [61] 0.072111946 0.067251842 0.062719293 0.058492222 0.054550042 0.050873552
#>  [67] 0.047444845 0.044247221 0.041265107 0.038483977 0.035890285 0.033471400
#>  [73] 0.031215540 0.029111717 0.027149685 0.025319887 0.023613411 0.022021946
#>  [79] 0.020537741 0.019153566 0.017862680 0.016658795 0.015536048 0.014488970
#>  [85] 0.013512462 0.012601767 0.011752450 0.010960375 0.010221682 0.009532775
#>  [91] 0.008890298 0.008291121 0.007732328 0.007211195 0.006725184 0.006271929
#>  [97] 0.005849222 0.005455004 0.005087355 0.004744485
#> 
#> $fit
#> $beta
#>               4.7445     4.4247    4.1265    3.8484    3.5890    3.3471
#> (Intercept) 20.76667 22.5514110 24.215870 25.768149 27.215810 28.565904
#> am           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000 -0.5791133 -1.119196 -1.622879 -2.092616 -2.530694
#>                3.1216    2.9112    2.7150    2.5320    2.3613    2.2022
#> (Intercept) 29.825006 30.999249 32.094351 33.115648 34.068112 34.956383
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.939247 -3.320265 -3.675603 -4.006993 -4.316049 -4.604275
#>                2.0538   1.9154    1.7863    1.6659    1.5536      1.4489
#> (Intercept) 35.784788 36.55736 37.277866 37.949811 38.420801 37.17006453
#> am           0.000000  0.00000  0.000000  0.000000  0.000000  0.00000000
#> carb         0.000000  0.00000  0.000000  0.000000  0.000000  0.00000000
#> cyl          0.000000  0.00000  0.000000  0.000000  0.000000  0.00000000
#> disp         0.000000  0.00000  0.000000  0.000000  0.000000  0.00000000
#> drat         0.000000  0.00000  0.000000  0.000000  0.000000  0.00000000
#> gear         0.000000  0.00000  0.000000  0.000000  0.000000  0.00000000
#> hp           0.000000  0.00000  0.000000  0.000000  0.000000  0.00000000
#> qsec         0.000000  0.00000  0.000000  0.000000  0.000000  0.06417816
#> vs           0.000000  0.00000  0.000000  0.000000  0.000000  0.00000000
#> wt          -4.873075 -5.12376 -5.357549 -5.575581 -5.728408 -5.69851106
#>                 1.3512     1.2602     1.1752     1.0960     1.0222     0.9533
#> (Intercept) 35.5532827 34.0460827 32.6401117 31.3288984 30.1060565 28.9656300
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.1471389  0.2244768  0.2966204  0.3639017  0.4266485  0.4851664
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.6598646 -5.6238375 -5.5902301 -5.5588877 -5.5296577 -5.5023977
#>                 0.8890     0.8291    0.7732     0.7211     0.6725     0.6272
#> (Intercept) 27.9020645 26.9101797 25.985145 25.1224538 24.3179055 23.5675810
#> am           0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> qsec         0.5397404  0.5906363  0.638102  0.6823686  0.7236518  0.7621527
#> vs           0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> wt          -5.4769750 -5.4532656 -5.431154 -5.4105330 -5.3913016 -5.3733664
#>                 0.5849     0.5455     0.5087     0.4744    0.4425    0.4127
#> (Intercept) 22.8678259 22.2152319 21.9649805 21.9649629 11.250130 11.247650
#> am           0.0000000  0.0000000  0.0000000  0.0000000  2.740951  2.741528
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> qsec         0.7980587  0.8315448  0.8443858  0.8443867  1.141460  1.141534
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> wt          -5.3566399 -5.3410407 -5.3350589 -5.3350584 -4.022020 -4.021739
#>                0.3848    0.3589    0.3347    0.3122    0.2911    0.2715
#> (Intercept) 11.245926 11.244727 11.243894 11.243315 11.242913 11.242633
#> am           2.741929  2.742207  2.742401  2.742535  2.742629  2.742694
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         1.141586  1.141621  1.141646  1.141664  1.141676  1.141684
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -4.021543 -4.021407 -4.021312 -4.021246 -4.021201 -4.021169
#>                  0.2532     0.2361     0.2202     0.2054     0.1915     0.1786
#> (Intercept) 11.86578377 12.7406138 13.5554104 14.0852442 10.6628226 10.6590038
#> am           2.82048723  2.9328712  3.0372768  3.1040554  3.1755243  3.1761542
#> carb        -0.09567561 -0.2313825 -0.3576603 -0.4393100 -0.4965134 -0.4964299
#> cyl          0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         1.10412113  1.0511180  1.0017746  0.9697833  1.2423913  1.2425272
#> vs           0.00000000  0.0000000  0.0000000  0.0000000 -1.4973019 -1.4972933
#> wt          -3.93700281 -3.8166173 -3.7046799 -3.6326551 -3.8518712 -3.8515952
#>                 0.1666      0.1554      0.1449      0.1351      0.1260
#> (Intercept) 10.6565908 -4.58324773 -4.57958790 -4.57604724 -4.57265038
#> am           3.1765933  3.17906730  3.17899762  3.17890538  3.17879599
#> carb        -0.4963783 -0.45590115 -0.45592363 -0.45594439 -0.45596346
#> cyl          0.0000000  1.50620973  1.50591330  1.50563677  1.50538010
#> disp         0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         0.0000000  2.77804433  2.77772220  2.77744334  2.77720338
#> gear         0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.0000000 -0.03612677 -0.03612234 -0.03611871 -0.03611578
#> qsec         1.2426059  1.11624879  1.11616675  1.11608030  1.11599140
#> vs          -1.4972385  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -3.8513919 -3.04103304 -3.04095351 -3.04086208 -3.04076216
#>                  0.1175      0.1096     0.1022      0.0953     0.0889
#> (Intercept) -4.56941547 -4.56635517 -4.5634775 -4.56078644 -4.5582830
#> am           3.17867408  3.17854360  3.1784078  3.17826959  3.1781311
#> carb        -0.45598092 -0.45599683 -0.4560113 -0.45602432 -0.4560361
#> cyl          1.50514298  1.50492491  1.5047252  1.50454319  1.5043779
#> disp         0.00000000  0.00000000  0.0000000  0.00000000  0.0000000
#> drat         2.77699826  2.77682424  2.7766778  2.77655583  2.7764553
#> gear         0.00000000  0.00000000  0.0000000  0.00000000  0.0000000
#> hp          -0.03611345 -0.03611165 -0.0361103 -0.03610933 -0.0361087
#> qsec         1.11590168  1.11581248  1.1157249  1.11563974  1.1155577
#> vs           0.00000000  0.00000000  0.0000000  0.00000000  0.0000000
#> wt          -3.04065665 -3.04054797 -3.0404382 -3.04032887 -3.0402214
#>                  0.0829     0.0773      0.0721      0.0673      0.0627
#> (Intercept) -4.55596528 -4.5538295 -4.55186999 -4.55007995 -4.54845152
#> am           3.17799434  3.1778607  3.17773141  3.17760736  3.17748920
#> carb        -0.45604659 -0.4560560 -0.45606435 -0.45607175 -0.45607827
#> cyl          1.50422840  1.5040938  1.50397303  1.50386517  1.50376924
#> disp         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> drat         2.77637366  2.7763084  2.77625736  2.77621859  2.77619030
#> gear         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> hp          -0.03610833 -0.0361082 -0.03610825 -0.03610846 -0.03610878
#> qsec         1.11547934  1.1154049  1.11533478  1.11526899  1.11520762
#> vs           0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> wt          -3.04011696 -3.0400163 -3.03991994 -3.03982849 -3.03974220
#>                    0.0585      0.0546      0.0509      0.0474      0.0442
#> (Intercept) -4.5466599224 -4.92232670 -4.92174573 -4.92092860 -4.91994022
#> am           3.1773774166  3.14564811  3.14574003  3.14580631  3.14584969
#> carb        -0.4560839823 -0.48423549 -0.48425891 -0.48428156 -0.48430324
#> cyl          1.5036842788  1.42324100  1.42298877  1.42274927  1.42252229
#> disp         0.0000000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         2.7761709388  2.74175101  2.74141466  2.74110652  2.74082648
#> gear         0.0000000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.0361092048 -0.03477606 -0.03476844 -0.03476176 -0.03475592
#> qsec         1.1151506795  1.18645123  1.18654356  1.18661470  1.18666825
#> vs          -0.0002367474 -0.51306956 -0.51346198 -0.51381258 -0.51412713
#> wt          -3.0397272978 -3.09144904 -3.09156250 -3.09164835 -3.09171045
#>                  0.0413      0.0385      0.0359     0.0335      0.0312
#> (Intercept) -4.91883210 -4.91764529 -4.91641270 -5.1955259 -5.19200264
#> am           3.14587294  3.14587875  3.14586972  2.9832921  2.98286842
#> carb        -0.48432382 -0.48434323 -0.48436142 -0.5356739 -0.53571043
#> cyl          1.42230774  1.42210556  1.42191566  1.4468048  1.44694438
#> disp         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> drat         2.74057375  2.74034704  2.74014483  2.6435928  2.64326104
#> gear         0.00000000  0.00000000  0.00000000  0.2645334  0.26486537
#> hp          -0.03475082 -0.03474639 -0.03474254 -0.0357189 -0.03572582
#> qsec         1.18670721  1.18673406  1.18675091  1.1676112  1.16742830
#> vs          -0.51441015 -0.51466527 -0.51489545 -0.4938715 -0.49349790
#> wt          -3.09175231 -3.09177711 -3.09178768 -3.0344040 -3.03440988
#>                  0.0291      0.0271      0.0253      0.0236     0.0220
#> (Intercept) -5.18924413 -5.18710679 -5.18547057 -5.18423592 -5.1833209
#> am           2.98253157  2.98226432  2.98205283  2.98188597  2.9817548
#> carb        -0.53573720 -0.53575680 -0.53577117 -0.53578171 -0.5357895
#> cyl          1.44706525  1.44717061  1.44726303  1.44734460  1.4474170
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> drat         2.64300802  2.64281651  2.64267274  2.64256578  2.6424870
#> gear         0.26512693  0.26533379  0.26549811  0.26562932  0.2657347
#> hp          -0.03573152 -0.03573623 -0.03574012 -0.03574334 -0.0357460
#> qsec         1.16728138  1.16716382  1.16707014  1.16699582  1.1669372
#> vs          -0.49318686 -0.49292767 -0.49271133 -0.49253036 -0.4923785
#> wt          -3.03441574 -3.03442149 -3.03442720 -3.03443294 -3.0344388
#>                  0.0205      0.0192        0.0179       0.0167       0.0155
#> (Intercept) -5.18265849 -5.18219418 -5.184305e+00 -4.773394301 -4.770618520
#> am           2.98165226  2.98157254  2.981511e+00  2.986473617  2.986539717
#> carb        -0.53579518 -0.53579940 -5.358025e-01 -0.413124166 -0.413037106
#> cyl          1.44748146  1.44753918  1.447591e+00  1.375511673  1.375212850
#> disp         0.00000000  0.00000000  3.911948e-06  0.003728077  0.003729192
#> drat         2.64242967  2.64238849  2.642932e+00  2.614832628  2.614582730
#> gear         0.26581978  0.26588894  2.659394e-01  0.207290650  0.207043147
#> hp          -0.03574821 -0.03575005 -3.575642e-02 -0.037967041 -0.037962326
#> qsec         1.16689112  1.16685525  1.166835e+00  1.192333828  1.192356374
#> vs          -0.49225078 -0.49214290 -4.920494e-01 -0.410973287 -0.411214110
#> wt          -3.03444493 -3.03445132 -3.034445e+00 -3.343804227 -3.343994552
#>                   0.0145       0.0135       0.0126       0.0118       0.0110
#> (Intercept) -4.768011177 -4.765565180 -4.763273563 -4.761129503 -4.759126323
#> am           2.986598774  2.986651247  2.986697575  2.986738177  2.986773453
#> carb        -0.412960140 -0.412892579 -0.412833767 -0.412783080 -0.412739926
#> cyl          1.374933096  1.374671559  1.374427404  1.374199818  1.373988008
#> disp         0.003730107  0.003730838  0.003731401  0.003731808  0.003732074
#> drat         2.614351336  2.614137485  2.613940247  2.613758726  2.613592055
#> gear         0.206813124  0.206599711  0.206402065  0.206219371  0.206050836
#> hp          -0.037957845 -0.037953590 -0.037949555 -0.037945732 -0.037942115
#> qsec         1.192375630  1.192391838  1.192405229  1.192416021  1.192424421
#> vs          -0.411444685 -0.411665202 -0.411875865 -0.412076893 -0.412268517
#> wt          -3.344159115 -3.344299780 -3.344418316 -3.344516401 -3.344595625
#>                   0.0102      0.0095       0.0089       0.0083       0.0077
#> (Intercept) -4.757257503 -4.75551669 -4.753897693 -4.752394504 -4.751001286
#> am           2.986803780  2.98682952  2.986851018  2.986868595  2.986882560
#> carb        -0.412703742 -0.41267400 -0.412650183 -0.412631827 -0.412618476
#> cyl          1.373791203  1.37360865  1.373439635  1.373283444  1.373139398
#> disp         0.003732211  0.00373223  0.003732142  0.003731958  0.003731687
#> drat         2.613439400  2.61329996  2.613172956  2.613057654  2.612953340
#> gear         0.205895699  0.20575322  0.205622695  0.205503435  0.205394784
#> hp          -0.037938696 -0.03793547 -0.037932425 -0.037929558 -0.037926861
#> qsec         1.192430626  1.19243482  1.192437175  1.192437857  1.192437021
#> vs          -0.412450976 -0.41262452 -0.412789401 -0.412945882 -0.413094225
#> wt          -3.344657491 -3.34470342 -3.344734752 -3.344752751 -3.344758608
#>                   0.0072       0.0067       0.0063       0.0058       0.0055
#> (Intercept) -4.749712385 -4.748522328 -4.747425826 -4.746417771 -4.745493241
#> am           2.986893204  2.986900802  2.986905614  2.986907884  2.986907843
#> carb        -0.412609706 -0.412605116 -0.412604331 -0.412606994 -0.412612774
#> cyl          1.373006842  1.372885142  1.372773686  1.372671887  1.372579181
#> disp         0.003731339  0.003730921  0.003730442  0.003729909  0.003729329
#> drat         2.612859333  2.612774981  2.612699662  2.612632781  2.612573770
#> gear         0.205296111  0.205206812  0.205126306  0.205054038  0.204989479
#> hp          -0.037924328 -0.037921950 -0.037919723 -0.037917638 -0.037915690
#> qsec         1.192434811  1.192431362  1.192426804  1.192421254  1.192414824
#> vs          -0.413234697 -0.413367565 -0.413493098 -0.413611561 -0.413723223
#> wt          -3.344753442 -3.344738303 -3.344714179 -3.344681993 -3.344642608
#>                   0.0051       0.0047
#> (Intercept) -4.744647497 -4.743875978
#> am           2.986905706  2.986901677
#> carb        -0.412621359 -0.412632457
#> cyl          1.372495024  1.372418897
#> disp         0.003728709  0.003728053
#> drat         2.612522089  2.612477223
#> gear         0.204932120  0.204881479
#> hp          -0.037913873 -0.037912180
#> qsec         1.192407618  1.192399733
#> vs          -0.413828345 -0.413927190
#> wt          -3.344596833 -3.344545419
#> 
#> $iter
#>   [1]   0   3   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   4
#>  [19]   4   3   3   3   3   3   3   3   3   3   3   3   3   3   3   1  76   1
#>  [37]   1   1   1   1   1   1  28  30  29  15  38   1   1 112   1   1   1   1
#>  [55]   1   1   1   1   1   1   1   1   1   1  31   1   1   1   1   1   1  33
#>  [73]   1   1   1   1   1   1   1   1   1  84   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.744484503 4.424722127 4.126510664 3.848397656 3.589028535 3.347140023
#>   [7] 3.121553988 2.911171697 2.714968468 2.531988680 2.361341117 2.202194629
#>  [13] 2.053774081 1.915356582 1.786267959 1.665879477 1.553604776 1.448897014
#>  [19] 1.351246204 1.260176732 1.175245038 1.096037456 1.022168200 0.953277484
#>  [25] 0.889029772 0.829112141 0.773232758 0.721119458 0.672518420 0.627192929
#>  [31] 0.584922225 0.545500425 0.508735522 0.474448450 0.442472213 0.412651066
#>  [37] 0.384839766 0.358902853 0.334714002 0.312155399 0.291117170 0.271496847
#>  [43] 0.253198868 0.236134112 0.220219463 0.205377408 0.191535658 0.178626796
#>  [49] 0.166587948 0.155360478 0.144889701 0.135124620 0.126017673 0.117524504
#>  [55] 0.109603746 0.102216820 0.095327748 0.088902977 0.082911214 0.077323276
#>  [61] 0.072111946 0.067251842 0.062719293 0.058492222 0.054550042 0.050873552
#>  [67] 0.047444845 0.044247221 0.041265107 0.038483977 0.035890285 0.033471400
#>  [73] 0.031215540 0.029111717 0.027149685 0.025319887 0.023613411 0.022021946
#>  [79] 0.020537741 0.019153566 0.017862680 0.016658795 0.015536048 0.014488970
#>  [85] 0.013512462 0.012601767 0.011752450 0.010960375 0.010221682 0.009532775
#>  [91] 0.008890298 0.008291121 0.007732328 0.007211195 0.006725184 0.006271929
#>  [97] 0.005849222 0.005455004 0.005087355 0.004744485
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 34
#> 
#> $loss
#>   [1] 650.72667 559.98010 484.05711 420.82446 368.44042 325.31588 290.08058
#>   [8] 261.55379 238.71890 220.70134 206.74944 196.21779 188.55287 183.28053
#>  [15] 179.99516 178.35033 178.01387 170.56473 161.80761 154.52975 148.51117
#>  [22] 143.56843 139.54173 136.29339 133.70494 131.67446 130.11439 128.94958
#>  [29] 128.11560 127.55729 127.22749 127.08593 127.07415 127.07415 113.88677
#>  [36] 113.88677 113.88677 113.88677 113.88677 113.88677 113.88677 113.88677
#>  [43] 112.52220 111.15834 110.49153 110.36947 107.30750 107.30749 107.30749
#>  [50]  92.43614  92.43613  92.43612  92.43612  92.43611  92.43611  92.43610
#>  [57]  92.43610  92.43610  92.43610  92.43609  92.43609  92.43609  92.43609
#>  [64]  92.43580  92.11877  92.11876  92.11875  92.11874  92.11874  92.11873
#>  [71]  92.11873  91.98513  91.98512  91.98511  91.98511  91.98511  91.98511
#>  [78]  91.98510  91.98510  91.98510  91.98477  91.82702  91.82701  91.82700
#>  [85]  91.82700  91.82699  91.82698  91.82698  91.82698  91.82697  91.82697
#>  [92]  91.82697  91.82697  91.82696  91.82696  91.82696  91.82696  91.82696
#>  [99]  91.82696  91.82696
#> 
#> $linear.predictors
#>      4.7445   4.4247   4.1265   3.8484   3.5890   3.3471   3.1216   2.9112
#> 1  20.76667 21.03413 21.28358 21.51621 21.73316 21.93549 22.12418 22.30015
#> 2  20.76667 20.88646 20.99818 21.10237 21.19954 21.29016 21.37467 21.45349
#> 3  20.76667 21.20787 21.61933 22.00307 22.36094 22.69469 23.00595 23.29623
#> 4  20.76667 20.68956 20.61765 20.55059 20.48805 20.42972 20.37533 20.32460
#> 5  20.76667 20.54768 20.34345 20.15299 19.97536 19.80970 19.65521 19.51113
#> 6  20.76667 20.72720 20.69040 20.65608 20.62407 20.59422 20.56638 20.54041
#> 7  20.76667 20.55926 20.36583 20.18544 20.01721 19.86032 19.71400 19.57754
#> 8  20.76667 20.19442 19.66074 19.16303 18.69886 18.26598 17.86227 17.48577
#> 9  20.76667 20.39132 20.04127 19.71481 19.41035 19.12642 18.86161 18.61466
#> 10 20.76667 19.41030 18.14535 16.96565 15.86546 14.83942 13.88253 12.99013
#> 11 20.76667 21.27736 21.75364 22.19781 22.61206 22.99838 23.35866 23.69467
#> 12 20.76667 21.48874 22.16214 22.79017 23.37586 23.92208 24.43149 24.90656
#> 13 20.76667 21.12390 21.45705 21.76775 22.05751 22.32774 22.57976 22.81480
#> 14 20.76667 20.56216 20.37143 20.19356 20.02767 19.87297 19.72869 19.59414
#> 15 20.76667 20.32762 19.91816 19.53629 19.18016 18.84804 18.53830 18.24943
#> 16 20.76667 20.32472 19.91256 19.52818 19.16970 18.83539 18.52360 18.23283
#> 17 20.76667 21.43083 22.05022 22.62788 23.16660 23.66901 24.13756 24.57454
#> 18 20.76667 21.31211 21.82079 22.29519 22.73761 23.15022 23.53502 23.89388
#> 19 20.76667 20.71562 20.66802 20.62362 20.58222 20.54360 20.50759 20.47401
#> 20 20.76667 20.94727 21.11570 21.27277 21.41926 21.55588 21.68329 21.80211
#> 21 20.76667 20.94148 21.10450 21.25654 21.39834 21.53057 21.65390 21.76891
#>      2.7150   2.5320   2.3613    2.2022    2.0538   1.9154    1.7863    1.6659
#> 1  22.46427 22.61733 22.76006 22.893184 23.017331 23.13311 23.241089 23.341789
#> 2  21.52699 21.59554 21.65947 21.719094 21.774697 21.82655 21.874914 21.920015
#> 3  23.56695 23.81942 24.05488 24.274466 24.479254 24.67024 24.848353 25.014463
#> 4  20.27729 20.23316 20.19202 20.153640 20.117852 20.08447 20.053347 20.024318
#> 5  19.37676 19.25145 19.13458 19.025593 18.923948 18.82915 18.740748 18.658301
#> 6  20.51620 20.49362 20.47256 20.452918 20.434601 20.41752 20.401588 20.386731
#> 7  19.45028 19.33159 19.22090 19.117679 19.021410 18.93163 18.847899 18.769812
#> 8  17.13465 16.80718 16.50179 16.216986 15.951372 15.70366 15.472643 15.257196
#> 9  18.38435 18.16956 17.96925 17.782439 17.608218 17.44574 17.294210 17.152894
#> 10 12.15788 11.38172 10.65786  9.982798  9.353228  8.76609  8.218523  7.707859
#> 11 24.00802 24.30026 24.57280 24.826979 25.064023 25.28509 25.491259 25.683533
#> 12 25.34962 25.76281 26.14816 26.507539 26.842695 27.15526 27.446764 27.718620
#> 13 23.03399 23.23841 23.42905 23.606846 23.772658 23.92729 24.071509 24.206004
#> 14 19.46865 19.35163 19.24248 19.140700 19.045775 18.95725 18.874687 18.797690
#> 15 17.98003 17.72879 17.49449 17.275969 17.072179 16.88212 16.704879 16.539580
#> 16 17.96166 17.70876 17.47290 17.252947 17.047814 16.85651 16.678092 16.511702
#> 17 24.98206 25.36212 25.71656 26.047112 26.355388 26.64289 26.911010 27.161062
#> 18 24.22856 24.54068 24.83177 25.103236 25.356407 25.59252 25.812712 26.018067
#> 19 20.44269 20.41348 20.38624 20.360833 20.337140 20.31504 20.294437 20.275219
#> 20 21.91293 22.01628 22.11266 22.202543 22.286370 22.36455 22.437456 22.505451
#> 21 21.87617 21.97621 22.06950 22.156500 22.237639 22.31331 22.383881 22.449696
#>       1.5536    1.4489    1.3512    1.2602    1.1752   1.0960   1.0222
#> 1  23.412373 23.296338 23.146344 23.006517 22.876080 22.75444 22.64099
#> 2  21.951629 21.879158 21.785477 21.698145 21.616679 21.54070 21.46985
#> 3  25.130895 25.143874 25.160652 25.176293 25.190883 25.20449 25.21718
#> 4  20.003970 20.096975 20.217199 20.329274 20.433822 20.53132 20.62225
#> 5  18.600510 18.750899 18.945300 19.126526 19.295580 19.45324 19.60027
#> 6  20.376317 20.689435 21.094191 21.471513 21.823494 22.15175 22.45789
#> 7  18.715078 18.741647 18.775991 18.808007 18.837873 18.86573 18.89170
#> 8  15.106181 15.093825 15.077851 15.062961 15.049070 15.03612 15.02403
#> 9  17.053840 17.044154 17.031633 17.019961 17.009072 16.99892 16.98945
#> 10  7.349917  7.404995  7.476193  7.542565  7.604479  7.66222  7.71607
#> 11 25.818304 25.882889 25.966376 26.044204 26.116804 26.18451 26.24766
#> 12 27.909173 27.990442 28.095496 28.193429 28.284785 28.36998 28.44944
#> 13 24.300276 24.407440 24.545966 24.675104 24.795568 24.90791 25.01269
#> 14 18.743720 18.705961 18.657151 18.611650 18.569204 18.52962 18.49270
#> 15 16.423715 16.276768 16.086814 15.909734 15.744548 15.59050 15.44683
#> 16 16.395073 16.353527 16.299822 16.249757 16.203055 16.15950 16.11888
#> 17 27.336332 27.356413 27.382370 27.406569 27.429142 27.45019 27.46983
#> 18 26.162009 26.047026 25.898393 25.759833 25.630580 25.51004 25.39762
#> 19 20.261748 20.036368 19.745026 19.473432 19.220078 18.98380 18.76345
#> 20 22.553112 22.379950 22.156111 21.947443 21.752790 21.57126 21.40196
#> 21 22.495827 22.521918 22.555643 22.587083 22.616411 22.64376 22.66927
#>       0.9533    0.8890    0.8291    0.7732    0.7211    0.6725    0.6272
#> 1  22.535188 22.436517 22.344497 22.258679 22.178644 22.104004 22.034394
#> 2  21.403769 21.342143 21.284671 21.231072 21.181085 21.134467 21.090991
#> 3  25.229015 25.240052 25.250345 25.259944 25.268897 25.277246 25.285032
#> 4  20.707057 20.786144 20.859900 20.928686 20.992836 21.052662 21.108456
#> 5  19.737399 19.865282 19.984546 20.095773 20.199502 20.296241 20.386460
#> 6  22.743389 23.009649 23.257964 23.489544 23.705516 23.906932 24.094773
#> 7  18.915928 18.938520 18.959590 18.979240 18.997565 19.014656 19.030594
#> 8  15.012767 15.002260 14.992460 14.983321 14.974798 14.966849 14.959436
#> 9  16.980616 16.972379 16.964698 16.957534 16.950853 16.944622 16.938811
#> 10  7.766291  7.813126  7.856806  7.897541  7.935531  7.970961  8.004002
#> 11 26.306546 26.361465 26.412684 26.460450 26.504998 26.546542 26.585287
#> 12 28.523542 28.592650 28.657099 28.717206 28.773261 28.825538 28.874292
#> 13 25.110400 25.201527 25.286512 25.365770 25.439685 25.508620 25.572908
#> 14 18.458273 18.426165 18.396220 18.368294 18.342250 18.317961 18.295309
#> 15 15.312838 15.187880 15.071345 14.962664 14.861307 14.766782 14.678627
#> 16 16.080999 16.045670 16.012722 15.981995 15.953339 15.926614 15.901690
#> 17 27.488136 27.505212 27.521137 27.535988 27.549839 27.562756 27.574802
#> 18 25.292778 25.195003 25.103817 25.018777 24.939469 24.865505 24.796526
#> 19 18.557943 18.366290 18.187554 18.020864 17.865409 17.720430 17.585223
#> 20 21.244068 21.096820 20.959496 20.831428 20.711991 20.600603 20.496722
#> 21 22.693060 22.715246 22.735936 22.755232 22.773228 22.790010 22.805662
#>       0.5849    0.5455    0.5087    0.4744    0.4425    0.4127    0.3848
#> 1  21.969476 21.908932 21.885716 21.885714 22.241819 22.241873 22.241910
#> 2  21.050445 21.012632 20.998132 20.998131 21.855422 21.855589 21.855705
#> 3  25.292294 25.299066 25.301663 25.301663 25.902564 25.902693 25.902782
#> 4  21.160490 21.209017 21.227625 21.227627 20.509317 20.509182 20.509089
#> 5  20.470599 20.549067 20.579157 20.579159 20.414260 20.414253 20.414248
#> 6  24.269954 24.433329 24.495979 24.495983 24.720199 24.720303 24.720375
#> 7  19.045459 19.059321 19.064637 19.064638 18.303098 18.302942 18.302834
#> 8  14.952523 14.946075 14.943603 14.943603 14.741911 14.741866 14.741835
#> 9  16.933392 16.928338 16.926400 16.926400 16.337690 16.337564 16.337477
#> 10  8.034817  8.063555  8.074575  8.074576  9.775509  9.775876  9.776132
#> 11 26.621421 26.655119 26.668042 26.668043 27.366862 27.367021 27.367131
#> 12 28.919760 28.962163 28.978424 28.978425 29.325727 29.325815 29.325876
#> 13 25.632863 25.688778 25.710219 25.710221 24.176464 24.176161 24.175950
#> 14 18.274183 18.254482 18.246927 18.246926 17.181748 17.181517 17.181356
#> 15 14.596413 14.519741 14.490339 14.490337 13.395470 13.395213 13.395035
#> 16 15.878446 15.856769 15.848456 15.848456 15.247355 15.247221 15.247127
#> 17 27.586037 27.596515 27.600532 27.600533 27.782065 27.782107 27.782136
#> 18 24.732197 24.672203 24.649197 24.649195 24.446339 24.446276 24.446232
#> 19 17.459129 17.341532 17.296437 17.296434 17.792446 17.792510 17.792554
#> 20 20.399843 20.309493 20.274847 20.274844 20.542714 20.542740 20.542757
#> 21 22.820259 22.833872 22.839092 22.839092 24.041020 24.041278 24.041457
#>      0.3589    0.3347    0.3122    0.2911   0.2715    0.2532    0.2361   0.2202
#> 1  22.24194 22.241955 22.241967 22.241976 22.24198 22.162455 22.049819 21.94500
#> 2  21.85579 21.855842 21.855881 21.855908 21.85593 21.776827 21.665208 21.56130
#> 3  25.90284 25.902888 25.902918 25.902939 25.90295 26.004443 26.148855 26.28320
#> 4  20.50902 20.508979 20.508948 20.508926 20.50891 20.576759 20.672540 20.76170
#> 5  20.41424 20.414242 20.414240 20.414239 20.41424 20.473408 20.557340 20.63544
#> 6  24.72043 24.720460 24.720485 24.720502 24.72051 24.557248 24.326105 24.11099
#> 7  18.30276 18.302707 18.302671 18.302646 18.30263 18.145208 17.921379 17.71315
#> 8  14.74181 14.741799 14.741788 14.741781 14.74178 14.766863 14.802286 14.83526
#> 9  16.33742 16.337374 16.337345 16.337324 16.33731 16.326268 16.310160 16.29521
#> 10  9.77631  9.776433  9.776519  9.776579  9.77662  9.804217  9.844673  9.88221
#> 11 27.36721 27.367261 27.367298 27.367324 27.36734 27.426428 27.510811 27.58928
#> 12 29.32592 29.325948 29.325968 29.325983 29.32599 29.338206 29.355857 29.37225
#> 13 24.17580 24.175702 24.175631 24.175582 24.17555 24.158860 24.134140 24.11122
#> 14 17.18124 17.181167 17.181113 17.181076 17.18105 17.252123 17.352109 17.44522
#> 15 13.39491 13.394824 13.394764 13.394723 13.39469 13.379497 13.357001 13.33615
#> 16 15.24706 15.247018 15.246986 15.246965 15.24695 15.361922 15.524516 15.67585
#> 17 27.78216 27.782170 27.782180 27.782187 27.78219 27.840384 27.923077 28.00001
#> 18 24.44620 24.446180 24.446165 24.446155 24.44615 24.508557 24.596829 24.67899
#> 19 17.79258 17.792606 17.792621 17.792631 17.79264 17.833026 17.890488 17.94394
#> 20 20.54277 20.542777 20.542783 20.542787 20.54279 20.320597 20.005488 19.71227
#> 21 24.04158 24.041668 24.041728 24.041770 24.04180 24.086705 24.151318 24.21136
#>       0.2054    0.1915    0.1786    0.1666    0.1554    0.1449    0.1351
#> 1  21.877137 22.210152 22.210257 22.210317 23.075850 23.075660 23.075495
#> 2  21.493888 21.923664 21.923915 21.924072 22.925486 22.925271 22.925080
#> 3  26.369897 26.029092 26.029165 26.029234 25.218631 25.218843 25.219032
#> 4  20.819536 20.437328 20.437131 20.437008 20.503496 20.503511 20.503526
#> 5  20.685966 20.462685 20.462661 20.462650 19.920777 19.920828 19.920869
#> 6  23.971799 24.489860 24.490199 24.490387 23.970522 23.970482 23.970435
#> 7  17.578705 16.664791 16.664751 16.664739 19.042946 19.042792 19.042676
#> 8  14.856638 15.113775 15.113695 15.113634 15.170228 15.170154 15.170086
#> 9  16.285697 16.671890 16.671743 16.671628 16.427429 16.427311 16.427195
#> 10  9.906022  9.923632  9.924067  9.924365  9.606693  9.606847  9.606992
#> 11 27.639829 27.559773 27.559930 27.560042 28.157902 28.157840 28.157782
#> 12 29.382755 29.499935 29.500049 29.500121 30.172919 30.172743 30.172572
#> 13 24.096803 24.034395 24.034068 24.033837 22.600149 22.600393 22.600623
#> 14 17.505705 17.931988 17.931635 17.931385 18.751607 18.751354 18.751126
#> 15 13.322970 13.030833 13.030503 13.030290 12.677698 12.677822 12.677936
#> 16 15.773871 16.042123 16.041849 16.041663 16.128089 16.128023 16.127964
#> 17 28.049707 27.872356 27.872362 27.872376 28.327514 28.327478 28.327444
#> 18 24.732179 25.350250 25.350089 25.349968 24.861600 24.861736 24.861884
#> 19 17.978401 17.656535 17.656526 17.656544 17.553305 17.553306 17.553278
#> 20 19.522627 19.446648 19.446832 19.446950 17.510201 17.510436 17.510640
#> 21 24.249868 23.748294 23.748576 23.748789 23.496956 23.497169 23.497366
#>       0.1260    0.1175    0.1096    0.1022    0.0953    0.0889    0.0829
#> 1  23.075352 23.075228 23.075122 23.075031 23.074954 23.074890 23.074837
#> 2  22.924913 22.924765 22.924637 22.924525 22.924429 22.924346 22.924276
#> 3  25.219200 25.219349 25.219480 25.219595 25.219696 25.219783 25.219859
#> 4  20.503540 20.503554 20.503568 20.503581 20.503593 20.503604 20.503615
#> 5  19.920901 19.920924 19.920942 19.920953 19.920960 19.920963 19.920962
#> 6  23.970384 23.970330 23.970275 23.970219 23.970164 23.970109 23.970056
#> 7  19.042594 19.042539 19.042509 19.042497 19.042502 19.042520 19.042548
#> 8  15.170023 15.169966 15.169914 15.169867 15.169825 15.169787 15.169753
#> 9  16.427080 16.426970 16.426863 16.426761 16.426665 16.426574 16.426489
#> 10  9.607128  9.607254  9.607371  9.607479  9.607579  9.607669  9.607752
#> 11 28.157727 28.157676 28.157629 28.157585 28.157545 28.157509 28.157476
#> 12 30.172406 30.172247 30.172095 30.171952 30.171818 30.171692 30.171575
#> 13 22.600838 22.601038 22.601224 22.601396 22.601554 22.601698 22.601829
#> 14 18.750921 18.750737 18.750573 18.750428 18.750299 18.750186 18.750086
#> 15 12.678041 12.678136 12.678223 12.678302 12.678374 12.678437 12.678495
#> 16 16.127912 16.127866 16.127826 16.127792 16.127762 16.127736 16.127714
#> 17 28.327414 28.327386 28.327361 28.327338 28.327318 28.327300 28.327284
#> 18 24.862040 24.862200 24.862362 24.862522 24.862680 24.862833 24.862981
#> 19 17.553225 17.553154 17.553067 17.552971 17.552866 17.552757 17.552646
#> 20 17.510816 17.510968 17.511098 17.511208 17.511300 17.511377 17.511440
#> 21 23.497547 23.497711 23.497861 23.497997 23.498119 23.498229 23.498327
#>       0.0773    0.0721    0.0673   0.0627    0.0585   0.0546    0.0509
#> 1  23.074793 23.074758 23.074730 23.07471 23.074836 23.12268 23.122493
#> 2  22.924216 22.924166 22.924125 22.92409 22.924190 22.99877 22.998609
#> 3  25.219923 25.219978 25.220025 25.22006 25.220023 25.14824 25.148361
#> 4  20.503625 20.503634 20.503643 20.50365 20.503525 20.36464 20.364528
#> 5  19.920959 19.920954 19.920947 19.92094 19.920781 19.82919 19.829189
#> 6  23.970006 23.969957 23.969911 23.96987 23.969698 24.16470 24.164999
#> 7  19.042584 19.042626 19.042673 19.04272 19.042626 18.71479 18.714289
#> 8  15.169723 15.169697 15.169673 15.16965 15.169682 15.23043 15.230410
#> 9  16.426410 16.426336 16.426269 16.42621 16.426219 16.51882 16.518850
#> 10  9.607827  9.607894  9.607955  9.60801  9.608016  9.64960  9.649729
#> 11 28.157446 28.157419 28.157394 28.15737 28.157287 28.10912 28.109049
#> 12 30.171467 30.171368 30.171277 30.17119 30.171076 30.16630 30.166249
#> 13 22.601948 22.602056 22.602153 22.60224 22.602233 22.66500 22.665219
#> 14 18.750000 18.749925 18.749860 18.74981 18.749848 18.82172 18.821523
#> 15 12.678546 12.678591 12.678631 12.67867 12.678760 12.64530 12.645374
#> 16 16.127695 16.127680 16.127667 16.12766 16.127710 16.19628 16.196237
#> 17 28.327269 28.327257 28.327246 28.32724 28.327179 28.25208 28.251983
#> 18 24.863122 24.863256 24.863382 24.86350 24.863786 25.12718 25.127304
#> 19 17.552534 17.552424 17.552315 17.55221 17.552216 17.46527 17.465399
#> 20 17.511491 17.511532 17.511563 17.51159 17.511738 17.52336 17.523615
#> 21 23.498415 23.498492 23.498561 23.49862 23.498570 23.38652 23.386590
#>       0.0474    0.0442    0.0413    0.0385   0.0359    0.0335    0.0312
#> 1  23.122328 23.122181 23.122049 23.121932 23.12183 23.033707 23.033746
#> 2  22.998462 22.998329 22.998208 22.998099 22.99800 22.913797 22.913731
#> 3  25.148472 25.148578 25.148676 25.148767 25.14885 25.148975 25.148960
#> 4  20.364428 20.364339 20.364262 20.364193 20.36413 20.405297 20.405634
#> 5  19.829188 19.829187 19.829184 19.829181 19.82918 19.905249 19.905583
#> 6  24.165247 24.165458 24.165637 24.165791 24.16592 24.234119 24.233665
#> 7  18.713848 18.713460 18.713121 18.712825 18.71257 18.805263 18.805661
#> 8  15.230388 15.230366 15.230346 15.230325 15.23031 15.218329 15.218386
#> 9  16.518871 16.518882 16.518883 16.518876 16.51886 16.483549 16.483571
#> 10  9.649854  9.649973  9.650087  9.650196  9.65030  9.629256  9.628972
#> 11 28.108981 28.108918 28.108859 28.108804 28.10875 28.089686 28.089625
#> 12 30.166193 30.166131 30.166064 30.165993 30.16592 30.105138 30.104961
#> 13 22.665425 22.665618 22.665800 22.665971 22.66613 22.556402 22.556244
#> 14 18.821346 18.821183 18.821032 18.820893 18.82077 18.847143 18.847439
#> 15 12.645438 12.645498 12.645553 12.645604 12.64565 12.480064 12.479781
#> 16 16.196195 16.196158 16.196126 16.196097 16.19607 16.233110 16.233300
#> 17 28.251897 28.251820 28.251750 28.251687 28.25163 28.228264 28.228309
#> 18 25.127431 25.127564 25.127702 25.127841 25.12798 25.292483 25.292562
#> 19 17.465498 17.465568 17.465612 17.465634 17.46564 17.579649 17.579483
#> 20 17.523843 17.524050 17.524235 17.524402 17.52455 17.588891 17.589006
#> 21 23.386666 23.386741 23.386815 23.386887 23.38696 23.321631 23.321382
#>      0.0291    0.0271    0.0253    0.0236    0.0220    0.0205   0.0192
#> 1  23.03378 23.033820 23.033854 23.033884 23.033911 23.033935 23.03396
#> 2  22.91369 22.913655 22.913634 22.913621 22.913614 22.913611 22.91361
#> 3  25.14894 25.148926 25.148907 25.148889 25.148872 25.148855 25.14884
#> 4  20.40590 20.406123 20.406298 20.406440 20.406554 20.406647 20.40672
#> 5  19.90585 19.906057 19.906224 19.906356 19.906462 19.906546 19.90661
#> 6  24.23329 24.232992 24.232746 24.232547 24.232385 24.232255 24.23215
#> 7  18.80599 18.806269 18.806500 18.806693 18.806854 18.806989 18.80710
#> 8  15.21843 15.218469 15.218500 15.218525 15.218545 15.218562 15.21858
#> 9  16.48359 16.483605 16.483619 16.483631 16.483642 16.483652 16.48366
#> 10  9.62874  9.628549  9.628392  9.628263  9.628155  9.628065  9.62799
#> 11 28.08958 28.089543 28.089517 28.089498 28.089484 28.089475 28.08947
#> 12 30.10482 30.104718 30.104637 30.104577 30.104532 30.104499 30.10448
#> 13 22.55611 22.555998 22.555901 22.555817 22.555743 22.555678 22.55562
#> 14 18.84768 18.847879 18.848043 18.848177 18.848289 18.848382 18.84846
#> 15 12.47956 12.479377 12.479233 12.479117 12.479023 12.478946 12.47888
#> 16 16.23345 16.233573 16.233670 16.233748 16.233810 16.233861 16.23390
#> 17 28.22835 28.228381 28.228410 28.228435 28.228456 28.228475 28.22849
#> 18 25.29262 25.292649 25.292669 25.292678 25.292679 25.292674 25.29266
#> 19 17.57935 17.579244 17.579162 17.579099 17.579053 17.579019 17.57900
#> 20 17.58909 17.589156 17.589202 17.589234 17.589255 17.589268 17.58928
#> 21 23.32118 23.321016 23.320882 23.320772 23.320681 23.320606 23.32054
#>       0.0179    0.0167    0.0155    0.0145    0.0135    0.0126    0.0118
#> 1  23.034020 23.125828 23.125830 23.125828 23.125824 23.125816 23.125805
#> 2  22.913665 22.940865 22.940831 22.940798 22.940767 22.940736 22.940707
#> 3  25.148735 25.090701 25.090779 25.090854 25.090926 25.090996 25.091063
#> 4  20.406782 20.545243 20.545193 20.545139 20.545084 20.545026 20.544968
#> 5  19.906389 19.886094 19.886034 19.885977 19.885922 19.885871 19.885822
#> 6  24.232177 24.260243 24.260309 24.260370 24.260428 24.260482 24.260533
#> 7  18.807247 18.767414 18.767060 18.766730 18.766423 18.766138 18.765875
#> 8  15.218305 14.972196 14.972107 14.972031 14.971968 14.971916 14.971875
#> 9  16.483383 16.347557 16.347536 16.347520 16.347508 16.347499 16.347494
#> 10  9.628176  9.707169  9.707306  9.707434  9.707553  9.707664  9.707767
#> 11 28.089530 28.034654 28.034556 28.034466 28.034384 28.034309 28.034240
#> 12 30.104579 30.143556 30.143489 30.143424 30.143360 30.143298 30.143237
#> 13 22.555442 22.582369 22.582702 22.583015 22.583308 22.583583 22.583841
#> 14 18.848534 18.842733 18.842544 18.842367 18.842200 18.842042 18.841894
#> 15 12.478880 12.489958 12.490179 12.490386 12.490580 12.490761 12.490931
#> 16 16.234162 16.399370 16.399341 16.399307 16.399270 16.399229 16.399186
#> 17 28.228563 28.242250 28.242190 28.242133 28.242079 28.242028 28.241979
#> 18 25.292927 25.258760 25.258769 25.258782 25.258800 25.258822 25.258847
#> 19 17.579196 17.609960 17.609827 17.609699 17.609575 17.609455 17.609340
#> 20 17.588779 17.604727 17.604965 17.605187 17.605393 17.605583 17.605759
#> 21 23.320530 23.248352 23.248454 23.248554 23.248651 23.248745 23.248837
#>       0.0110    0.0102    0.0095    0.0089    0.0083    0.0077    0.0072
#> 1  23.125792 23.125778 23.125761 23.125743 23.125723 23.125702 23.125680
#> 2  22.940678 22.940651 22.940625 22.940600 22.940576 22.940553 22.940531
#> 3  25.091127 25.091189 25.091249 25.091306 25.091361 25.091413 25.091463
#> 4  20.544908 20.544847 20.544785 20.544724 20.544662 20.544600 20.544539
#> 5  19.885775 19.885732 19.885690 19.885651 19.885614 19.885579 19.885546
#> 6  24.260580 24.260623 24.260664 24.260702 24.260737 24.260769 24.260799
#> 7  18.765631 18.765407 18.765200 18.765010 18.764837 18.764678 18.764533
#> 8  14.971844 14.971822 14.971809 14.971803 14.971803 14.971811 14.971823
#> 9  16.347492 16.347492 16.347495 16.347500 16.347507 16.347516 16.347526
#> 10  9.707863  9.707951  9.708033  9.708109  9.708178  9.708242  9.708301
#> 11 28.034177 28.034121 28.034070 28.034024 28.033982 28.033946 28.033913
#> 12 30.143179 30.143121 30.143066 30.143012 30.142960 30.142910 30.142862
#> 13 22.584082 22.584307 22.584516 22.584711 22.584892 22.585060 22.585216
#> 14 18.841755 18.841625 18.841503 18.841389 18.841282 18.841183 18.841090
#> 15 12.491090 12.491238 12.491375 12.491503 12.491622 12.491731 12.491833
#> 16 16.399140 16.399092 16.399042 16.398991 16.398939 16.398886 16.398833
#> 17 28.241933 28.241889 28.241848 28.241809 28.241772 28.241737 28.241704
#> 18 25.258876 25.258907 25.258940 25.258976 25.259013 25.259052 25.259091
#> 19 17.609230 17.609124 17.609022 17.608925 17.608832 17.608742 17.608657
#> 20 17.605922 17.606072 17.606210 17.606337 17.606453 17.606558 17.606654
#> 21 23.248926 23.249013 23.249097 23.249177 23.249256 23.249331 23.249404
#>       0.0067    0.0063    0.0058    0.0055    0.0051    0.0047
#> 1  23.125657 23.125634 23.125610 23.125585 23.125561 23.125536
#> 2  22.940511 22.940491 22.940472 22.940454 22.940437 22.940420
#> 3  25.091510 25.091555 25.091599 25.091640 25.091679 25.091716
#> 4  20.544478 20.544418 20.544359 20.544301 20.544243 20.544187
#> 5  19.885515 19.885486 19.885459 19.885434 19.885410 19.885388
#> 6  24.260827 24.260852 24.260875 24.260896 24.260915 24.260933
#> 7  18.764402 18.764284 18.764177 18.764081 18.763995 18.763920
#> 8  14.971841 14.971863 14.971889 14.971919 14.971952 14.971988
#> 9  16.347538 16.347551 16.347565 16.347581 16.347596 16.347613
#> 10  9.708354  9.708403  9.708448  9.708488  9.708524  9.708556
#> 11 28.033885 28.033860 28.033838 28.033820 28.033804 28.033791
#> 12 30.142815 30.142770 30.142727 30.142686 30.142646 30.142608
#> 13 22.585360 22.585492 22.585614 22.585726 22.585828 22.585922
#> 14 18.841004 18.840924 18.840850 18.840782 18.840719 18.840661
#> 15 12.491926 12.492012 12.492091 12.492164 12.492230 12.492290
#> 16 16.398779 16.398725 16.398672 16.398618 16.398565 16.398513
#> 17 28.241674 28.241645 28.241618 28.241592 28.241569 28.241547
#> 18 25.259132 25.259174 25.259216 25.259258 25.259300 25.259343
#> 19 17.608576 17.608499 17.608426 17.608356 17.608290 17.608227
#> 20 17.606741 17.606820 17.606891 17.606954 17.607011 17.607061
#> 21 23.249473 23.249540 23.249605 23.249667 23.249726 23.249782
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 22.8 21.4 18.1 22.8 19.2 16.4 17.3 10.4 32.4 33.9 21.5 15.2 13.3
#> [16] 19.2 27.3 26.0 15.8 19.7 21.4
#> 
#> $X
#>               am       carb         cyl       disp        drat       gear
#>  [1,]  1.0488088  1.0287711  0.05643326 -0.4311053  0.49165446  0.4082483
#>  [2,]  1.0488088  1.0287711  0.05643326 -0.4311053  0.49165446  0.4082483
#>  [3,]  1.0488088 -1.0619573 -1.12866530 -0.8966582  0.38861311  0.4082483
#>  [4,] -0.9534626 -1.0619573  0.05643326  0.4462827 -1.19822375 -1.0206207
#>  [5,] -0.9534626 -1.0619573  0.05643326  0.1508357 -1.85768842 -1.0206207
#>  [6,] -0.9534626 -0.3650478 -1.12866530 -0.6030018  0.53287101  0.4082483
#>  [7,] -0.9534626  1.0287711  0.05643326 -0.3630630  0.53287101  0.4082483
#>  [8,] -0.9534626  0.3318617  1.24153183  0.6056450 -1.21883202 -1.0206207
#>  [9,] -0.9534626  0.3318617  1.24153183  0.6056450 -1.21883202 -1.0206207
#> [10,] -0.9534626  1.0287711  1.24153183  2.2547764 -1.36308992 -1.0206207
#> [11,]  1.0488088 -1.0619573 -1.12866530 -1.1589793  0.86260334  0.4082483
#> [12,]  1.0488088 -1.0619573 -1.12866530 -1.2270216  1.15111913  0.4082483
#> [13,] -0.9534626 -1.0619573 -1.12866530 -0.7883276  0.07948905 -1.0206207
#> [14,] -0.9534626 -0.3650478  1.24153183  0.8581179 -1.05396586 -1.0206207
#> [15,] -0.9534626  1.0287711  1.24153183  1.2699531  0.14131386 -1.0206207
#> [16,] -0.9534626 -0.3650478  1.24153183  1.7176001 -1.19822375 -1.0206207
#> [17,]  1.0488088 -1.0619573 -1.12866530 -1.1562934  0.86260334  0.4082483
#> [18,]  1.0488088 -0.3650478 -1.12866530 -0.7865370  1.58389282  1.8371173
#> [19,]  1.0488088  1.0287711  1.24153183  1.2789061  1.15111913  1.8371173
#> [20,]  1.0488088  2.4225901  0.05643326 -0.5653994 -0.08537712  1.8371173
#> [21,]  1.0488088 -0.3650478 -1.12866530 -0.7802700  0.92442815  0.4082483
#>               hp       qsec         vs         wt
#>  [1,] -0.4321318 -0.8452793 -0.9534626 -0.5576383
#>  [2,] -0.4321318 -0.5481049 -0.9534626 -0.2497557
#>  [3,] -0.7322700  0.2956582  1.0488088 -0.9198530
#>  [4,] -0.4321318  0.7361132  1.0488088  0.1607544
#>  [5,] -0.5204077  1.1500347  1.0488088  0.4565631
#>  [6,] -0.6969596  2.5722267  1.0488088  0.0822745
#>  [7,] -0.2026143  0.1311510  1.0488088  0.4324154
#>  [8,]  0.8037315 -0.3464508 -0.9534626  1.1930665
#>  [9,]  0.8037315 -0.2403171 -0.9534626  0.7825564
#> [10,]  1.4216631 -0.1235700 -0.9534626  2.8278624
#> [11,] -1.2089601  0.7520333  1.0488088 -1.0647389
#> [12,] -1.2266153  0.9802208  1.0488088 -1.5054336
#> [13,] -0.6616492  1.0385943  1.0488088 -0.7447826
#> [14,]  0.2740758 -0.3995177 -0.9534626  0.4263785
#> [15,]  1.9513188 -1.4024814 -0.9534626  0.9153685
#> [16,]  0.7154555 -0.5321848 -0.9534626  0.9214054
#> [17,] -1.2089601  0.4495521  1.0488088 -1.3846953
#> [18,] -0.7675804 -0.7179188 -0.9534626 -1.1371819
#> [19,]  2.2867674 -1.8853898 -0.9534626  0.1064221
#> [20,]  0.7154555 -1.3547212 -0.9534626 -0.3765309
#> [21,] -0.4497870  0.2903516  1.0488088 -0.3644570
#> attr(,"center")
#>  [1]   0.4761905   2.5238095   5.9047619 208.1523810   3.6614286   3.7142857
#>  [7] 134.4761905  18.0528571   0.4761905   3.0818571
#> attr(,"scale")
#>  [1]   0.4994328   1.4349066   1.6876233 111.6951618   0.4852421   0.6998542
#>  [7]  56.6405702   1.8844151   0.4994328   0.8282379
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 30
#> 
#> $lambda.min
#> [1] 0.6271929
#> 
#> $null.dev
#> [1] 30.98698
#> 
#> $Bias
#> [1] 5.529835
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 6.741175