Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 46.844214 44.018570 38.232727 33.502068 29.668830 26.597168 24.170208
#>   [8] 22.287491 20.862753 19.822003 19.101855 18.648077 18.414345 18.361151
#>  [15] 18.454862 18.582671 18.191192 17.942301 17.903995 17.792052 17.668796
#>  [22] 17.598818 17.531758 17.343967 16.931583 16.571667 14.891985 14.382563
#>  [29] 13.556236 12.938452 12.693510 12.488517  9.912388  9.058505  8.974750
#>  [36]  8.020570  8.752624  9.192778  8.986603  8.947031  8.947193  8.947011
#>  [43]  8.948865  9.122104 12.803213 12.754390 14.795595 13.728653 14.962513
#>  [50] 14.962482 14.836550 14.855423 15.297112 15.258872 15.000312 25.106907
#>  [57] 25.013473 31.047848 34.669372 49.638296 49.642325 49.646299 49.545722
#>  [64] 49.549569 49.585784 49.485587 49.489291 57.108793 57.114117 57.150897
#>  [71] 57.204106 57.240715 57.245576 57.488214 57.493047 57.497774 57.502397
#>  [78] 57.506597 57.580696 57.518736 57.752056 62.955735 62.960730 62.965529
#>  [85] 62.970135 62.974558 62.630489 62.633174 61.764760 61.765491 61.766199
#>  [92] 61.766884 61.767548 61.768191 61.768815 61.769413 61.775456 61.776048
#>  [99] 61.776624 61.777156
#> 
#> $cvse
#>   [1] 14.079585 13.853388 12.241834 10.816061  9.551886  8.430914  7.440000
#>   [8]  6.570829  5.819581  5.186456  4.674730  4.288843  4.031244  3.898726
#>  [15]  3.880186  3.947080  3.882914  3.815050  3.817503  3.816081  3.792726
#>  [22]  3.774256  3.763879  3.753914  3.754664  3.745152  3.712626  3.670573
#>  [29]  3.593746  3.546345  3.530576  3.516603  2.839110  2.185906  2.170092
#>  [36]  1.943243  2.229144  2.542654  2.539675  2.530072  2.528686  2.528712
#>  [43]  2.528448  2.606747  5.822592  5.818685  6.017843  5.947281  6.047312
#>  [50]  6.047308  5.974303  5.971948  5.965537  5.966859  5.967686 12.971329
#>  [57] 12.968068 13.874267 14.056414 24.443420 24.446758 24.450046 24.461962
#>  [64] 24.465151 24.461937 24.475216 24.478261 26.813902 26.817286 26.819983
#>  [71] 26.820035 26.821188 26.824319 26.806101 26.809096 26.812033 26.814914
#>  [78] 26.817759 26.821214 26.803047 26.837722 29.546839 29.549208 29.551456
#>  [85] 29.553589 29.555609 29.499474 29.501164 28.860720 28.861026 28.861332
#>  [92] 28.861636 28.861939 28.862240 28.862540 28.862839 28.862623 28.862916
#>  [99] 28.863209 28.863502
#> 
#> $fold
#>  [1]  1  4  7  2  8  4  9  5  8  5  3 10  2  1 10  9  6  7  6  1  3
#> 
#> $lambda
#>   [1] 5.970626570 5.568226322 5.192946504 4.842959290 4.516560042 4.212159012
#>   [7] 3.928273592 3.663521099 3.416612039 3.186343822 2.971594913 2.771319361
#>  [13] 2.584541712 2.410352251 2.247902576 2.096401466 1.955111024 1.823343084
#>  [19] 1.700455863 1.585850829 1.478969791 1.379292177 1.286332500 1.199637995
#>  [25] 1.118786409 1.043383950 0.973063363 0.907482150 0.846320890 0.789281694
#>  [31] 0.736086750 0.686476967 0.640210717 0.597062657 0.556822632 0.519294650
#>  [37] 0.484295929 0.451656004 0.421215901 0.392827359 0.366352110 0.341661204
#>  [43] 0.318634382 0.297159491 0.277131936 0.258454171 0.241035225 0.224790258
#>  [49] 0.209640147 0.195511102 0.182334308 0.170045586 0.158585083 0.147896979
#>  [55] 0.137929218 0.128633250 0.119963799 0.111878641 0.104338395 0.097306336
#>  [61] 0.090748215 0.084632089 0.078928169 0.073608675 0.068647697 0.064021072
#>  [67] 0.059706266 0.055682263 0.051929465 0.048429593 0.045165600 0.042121590
#>  [73] 0.039282736 0.036635211 0.034166120 0.031863438 0.029715949 0.027713194
#>  [79] 0.025845417 0.024103523 0.022479026 0.020964015 0.019551110 0.018233431
#>  [85] 0.017004559 0.015858508 0.014789698 0.013792922 0.012863325 0.011996380
#>  [91] 0.011187864 0.010433839 0.009730634 0.009074821 0.008463209 0.007892817
#>  [97] 0.007360867 0.006864770 0.006402107 0.005970627
#> 
#> $fit
#> $beta
#>               5.9706     5.5682     5.1929    4.8430    4.5166    4.2122
#> (Intercept) 19.71905 22.1236206 24.3661334 26.457508 28.407932 30.226903
#> am           0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> cyl          0.00000 -0.3659133 -0.7071652 -1.025418 -1.322221 -1.599021
#> disp         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> wt           0.00000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#>                3.9283    3.6635    3.4166    3.1863    2.9716    2.7713
#> (Intercept) 31.923282 33.505330 34.980754 36.356739 37.639987 38.836749
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.857166 -2.097913 -2.322434 -2.531823 -2.727099 -2.909215
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                2.5845    2.4104    2.2479    2.0964   1.9551   1.8233   1.7005
#> (Intercept) 39.952853 40.993735 41.964465 42.869771 43.50433 43.50433 43.50433
#> am           0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#> carb         0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#> cyl         -3.079057 -3.237452 -3.385172 -3.522936 -3.61950 -3.61950 -3.61950
#> disp         0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#> drat         0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#> gear         0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#> hp           0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#> vs           0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#> wt           0.000000  0.000000  0.000000  0.000000  0.00000  0.00000  0.00000
#>               1.5859   1.4790   1.3793   1.2863   1.1996     1.1188    1.0434
#> (Intercept) 43.50433 43.50433 43.50433 43.50433 43.50433 43.5351072 43.580924
#> am           0.00000  0.00000  0.00000  0.00000  0.00000  0.0000000  0.000000
#> carb         0.00000  0.00000  0.00000  0.00000  0.00000  0.0000000  0.000000
#> cyl         -3.61950 -3.61950 -3.61950 -3.61950 -3.61950 -3.1910589 -2.508906
#> disp         0.00000  0.00000  0.00000  0.00000  0.00000  0.0000000  0.000000
#> drat         0.00000  0.00000  0.00000  0.00000  0.00000  0.0000000  0.000000
#> gear         0.00000  0.00000  0.00000  0.00000  0.00000  0.0000000  0.000000
#> hp           0.00000  0.00000  0.00000  0.00000  0.00000  0.0000000  0.000000
#> qsec         0.00000  0.00000  0.00000  0.00000  0.00000  0.0000000  0.000000
#> vs           0.00000  0.00000  0.00000  0.00000  0.00000  0.0000000  0.000000
#> wt           0.00000  0.00000  0.00000  0.00000  0.00000 -0.8425944 -2.183210
#>                0.9731    0.9075      0.8463     0.7893    0.7361     0.6865
#> (Intercept) 43.602186 43.601622 43.64550677 43.6969534 43.745147 43.7900957
#> am           0.000000  0.000000  0.00000000  0.0000000  0.000000  0.0000000
#> carb         0.000000  0.000000 -0.07867563 -0.1671823 -0.249676 -0.3266092
#> cyl         -2.183704 -2.183478 -2.16858074 -2.1526555 -2.137909 -2.1241585
#> disp         0.000000  0.000000  0.00000000  0.0000000  0.000000  0.0000000
#> drat         0.000000  0.000000  0.00000000  0.0000000  0.000000  0.0000000
#> gear         0.000000  0.000000  0.00000000  0.0000000  0.000000  0.0000000
#> hp           0.000000  0.000000  0.00000000  0.0000000  0.000000  0.0000000
#> qsec         0.000000  0.000000  0.00000000  0.0000000  0.000000  0.0000000
#> vs           0.000000  0.000000  0.00000000  0.0000000  0.000000  0.0000000
#> wt          -2.822149 -2.822422 -2.80006645 -2.7739119 -2.749417 -2.7265715
#>                 0.6402    0.5971     0.5568     0.5193     0.4843    0.4517
#> (Intercept) 43.8320148 43.871109 43.9075680 43.9415699 43.9727850 36.246286
#> am           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  5.592609
#> carb        -0.3983574 -0.465270 -0.5276729 -0.5858701 -0.6402424 -1.748859
#> cyl         -2.1113344 -2.099375 -2.0882209 -2.0778188 -2.0678791 -2.063670
#> disp         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.000000
#> drat         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.000000
#> gear         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.000000
#> hp           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.000000
#> qsec         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.000000
#> vs           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.000000
#> wt          -2.7052657 -2.685396 -2.6668652 -2.6495835 -2.6337047  0.000000
#>                0.4212    0.3928    0.3664    0.3417    0.3186    0.2972
#> (Intercept) 36.247141 36.247633 36.247917 36.248080 36.248175 36.248229
#> am           5.591973  5.591606  5.591395  5.591273  5.591203  5.591162
#> carb        -1.748708 -1.748621 -1.748571 -1.748543 -1.748526 -1.748516
#> cyl         -2.063831 -2.063924 -2.063977 -2.064008 -2.064026 -2.064036
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                 0.2771    0.2585     0.2410     0.2248     0.2096     0.1955
#> (Intercept) 57.9822134 57.988937 51.7139599 51.7075250 51.7018719 51.6969053
#> am           3.1000078  3.098971  2.4874435  2.4884850  2.4894037  2.4902131
#> carb        -1.9347093 -1.934711 -2.0240921 -2.0240884 -2.0240844 -2.0240804
#> cyl         -2.9499171 -2.950262 -2.7100836 -2.7097510 -2.7094590 -2.7092026
#> disp         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.000000  1.2899449  1.2899061  1.2898693  1.2898352
#> gear         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec        -0.8171506 -0.817381 -0.7822091 -0.7819832 -0.7817842 -0.7816092
#> vs           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#>                 0.1823     0.1700     0.1586     0.1479     0.1379     0.1286
#> (Intercept) 51.6925414 51.6887071 51.6853379 60.3363310 60.3410832 60.3455266
#> am           2.4909256  2.4915525  2.4921039  1.1663387  1.1655200  1.1647546
#> carb        -2.0240767 -2.0240732 -2.0240701 -2.2315955 -2.2316478 -2.2316966
#> cyl         -2.7089774 -2.7087795 -2.7086057 -3.5599880 -3.5603562 -3.5607005
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         1.2898042  1.2897764  1.2897515  1.5246999  1.5248102  1.5249132
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec        -0.7814553 -0.7813199 -0.7812009 -0.8851424 -0.8852564 -0.8853629
#> vs           0.0000000  0.0000000  0.0000000 -2.6549382 -2.6556456 -2.6563069
#> wt           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#>                  0.1200       0.1119      0.1043       0.0973      0.0907
#> (Intercept) 63.25028607 63.248914417 63.24769603 63.246614317 68.45843600
#> am           0.00000000  0.000000000  0.00000000  0.000000000  1.05276370
#> carb        -2.25083674 -2.250842685 -2.25084596 -2.250847443 -3.01875630
#> cyl         -3.50616650 -3.506311417 -3.50641203 -3.506478969 -3.57640205
#> disp        -0.00568336 -0.005683593 -0.00568384 -0.005684091 -0.01923179
#> drat         1.70642912  1.706542525  1.70663924  1.706722121  2.08000259
#> gear         0.00000000  0.000000000  0.00000000  0.000000000  0.00000000
#> hp           0.00000000  0.000000000  0.00000000  0.000000000  0.00000000
#> qsec        -0.98669022 -0.986558902 -0.98645392 -0.986370042 -1.47605965
#> vs          -3.13890796 -3.139795037 -3.14047201 -3.140984689 -3.25973790
#> wt           0.00000000  0.000000000  0.00000000  0.000000000  2.36347876
#>                  0.0846      0.0789      0.0736     0.0686      0.0640
#> (Intercept) 68.45982862 68.46115303 68.46241239 68.4636097 68.46474794
#> am           1.05274201  1.05272233  1.05270451  1.0526884  1.05267403
#> carb        -3.01896060 -3.01915617 -3.01934335 -3.0195225 -3.01969397
#> cyl         -3.57637219 -3.57634282 -3.57631394 -3.5762856 -3.57625771
#> disp        -0.01923621 -0.01924045 -0.01924451 -0.0192484 -0.01925212
#> drat         2.08023396  2.08045553  2.08066771  2.0808709  2.08106542
#> gear         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> qsec        -1.47620896 -1.47635174 -1.47648826 -1.4766188 -1.47674358
#> vs          -3.25974624 -3.25975306 -3.25975846 -3.2597625 -3.25976542
#> wt           2.36406465  2.36462569  2.36516291  2.3656773  2.36616977
#>                  0.0597      0.0557      0.0519      0.0484      0.0452
#> (Intercept) 68.46582979 68.46685792 68.46783483 68.46876294 68.46964455
#> am           1.05266115  1.05264971  1.05263962  1.05263078  1.05262312
#> carb        -3.01985807 -3.02001510 -3.02016536 -3.02030913 -3.02044670
#> cyl         -3.57623037 -3.57620355 -3.57617725 -3.57615149 -3.57612624
#> disp        -0.01925569 -0.01925911 -0.01926239 -0.01926552 -0.01926853
#> drat         2.08125168  2.08143001  2.08160073  2.08176416  2.08192061
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec        -1.47686288 -1.47697693 -1.47708595 -1.47719014 -1.47728973
#> vs          -3.25976716 -3.25976786 -3.25976759 -3.25976644 -3.25976448
#> wt           2.36664126  2.36709264  2.36752473  2.36793834  2.36833424
#>                 0.0421      0.0393     0.0366      0.0342      0.0319
#> (Intercept) 68.4704819 68.47127694 68.4720318 68.47274838 68.47342847
#> am           1.0526165  1.05261100  1.0526064  1.05260268  1.05259979
#> carb        -3.0205783 -3.02070422 -3.0208247 -3.02093989 -3.02105010
#> cyl         -3.5761015 -3.57607734 -3.5760537 -3.57603054 -3.57600792
#> disp        -0.0192714 -0.01927416 -0.0192768 -0.01927932 -0.01928174
#> drat         2.0820704  2.08221369  2.0823509  2.08248216  2.08260780
#> gear         0.0000000  0.00000000  0.0000000  0.00000000  0.00000000
#> hp           0.0000000  0.00000000  0.0000000  0.00000000  0.00000000
#> qsec        -1.4773849 -1.47747586 -1.4775628 -1.47764582 -1.47772516
#> vs          -3.2597618 -3.25975836 -3.2597543 -3.25974974 -3.25974462
#> wt           2.3687132  2.36907583  2.3694229  2.36975503  2.37007286
#>                  0.0297      0.0277      0.0258      0.0241     0.0225
#> (Intercept) 68.47407381 68.47468608 68.45553252 68.45722480 68.4588757
#> am           1.05259765  1.05259621  0.78998530  0.78957022  0.7891649
#> carb        -3.02115551 -3.02125632 -3.16200585 -3.16229442 -3.1625763
#> cyl         -3.57598581 -3.57596422 -3.61356595 -3.61366230 -3.6137562
#> disp        -0.01928406 -0.01928628 -0.02042667 -0.02043038 -0.0204340
#> drat         2.08272803  2.08284308  2.14469265  2.14489937  2.1451013
#> gear         0.00000000  0.00000000  0.27927333  0.27958458  0.2798887
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> qsec        -1.47780096 -1.47787338 -1.52595393 -1.52611662 -1.5262755
#> vs          -3.25973903 -3.25973302 -3.49058256 -3.49095238 -3.4913134
#> wt           2.37037697  2.37066795  2.59802794  2.59863151  2.5992212
#>                  0.0210     0.0196      0.0182      0.0170       0.0159
#> (Intercept) 68.46048630 68.4620576 68.46359054 68.46516084 68.653907715
#> am           0.78876922  0.7883828  0.78800556  0.78714464  0.607495683
#> carb        -3.16285178 -3.1631209 -3.16338375 -3.16380766 -3.215604484
#> cyl         -3.61384782 -3.6139371 -3.61402416 -3.61415545 -3.596852784
#> disp        -0.02043754 -0.0204410 -0.02044437 -0.02045083 -0.020337354
#> drat         2.14529867  2.1454915  2.14567981  2.14608414  2.239435956
#> gear         0.28018586  0.2804762  0.28075989  0.28132486  0.476710747
#> hp           0.00000000  0.0000000  0.00000000  0.00000000 -0.003957078
#> qsec        -1.52643072 -1.5265823 -1.52673035 -1.52696427 -1.588305949
#> vs          -3.49166586 -3.4920099 -3.49234583 -3.49286600 -3.597735046
#> wt           2.59979735  2.6003602  2.60091018  2.60190199  2.773471235
#>                   0.0148       0.0138       0.0129      0.0120       0.0112
#> (Intercept) 68.649001872 68.644160587 68.639384908 68.63467576 68.630033940
#> am           0.607599162  0.607724856  0.607871531  0.60803800  0.608223135
#> carb        -3.215465953 -3.215319749 -3.215166337 -3.21500617 -3.214839680
#> cyl         -3.596460071 -3.596075086 -3.595697794 -3.59532815 -3.594966122
#> disp        -0.020334040 -0.020330723 -0.020327406 -0.02032409 -0.020320776
#> drat         2.239468521  2.239488958  2.239497824  2.23949566  2.239482982
#> gear         0.477046069  0.477351396  0.477627936  0.47787686  0.478099301
#> hp          -0.003965719 -0.003973904 -0.003981649 -0.00398897 -0.003995882
#> qsec        -1.588218066 -1.588124210 -1.588024743 -1.58792002 -1.587810362
#> vs          -3.597246748 -3.596753618 -3.596256322 -3.59575549 -3.595251715
#> wt           2.773263246  2.773034261  2.772785424  2.77251784  2.772232566
#>                   0.0104       0.0097       0.0091       0.0085       0.0079
#> (Intercept) 68.625460146 68.620954962 68.616518874 68.612152273 68.607855463
#> am           0.608425831  0.608645045  0.608879768  0.609129034  0.609391915
#> carb        -3.214667286 -3.214489392 -3.214306386 -3.214118642 -3.213926519
#> cyl         -3.594611644 -3.594264663 -3.593925117 -3.593592941 -3.593268062
#> disp        -0.020317468 -0.020314168 -0.020310876 -0.020307595 -0.020304326
#> drat         2.239460301  2.239428100  2.239386854  2.239337017  2.239279029
#> gear         0.478296357  0.478469093  0.478618538  0.478745688  0.478851511
#> hp          -0.004002398 -0.004008535 -0.004014304 -0.004019719 -0.004024794
#> qsec        -1.587696107 -1.587577559 -1.587455019 -1.587328771 -1.587199092
#> vs          -3.594745573 -3.594237604 -3.593728320 -3.593218210 -3.592707734
#> wt           2.771930641  2.771613056  2.771280769  2.770934706  2.770575759
#>                  0.0074      0.0069       0.0064       0.0060
#> (Intercept) 68.60362866 68.59947201 68.595385582 68.591369363
#> am           0.60966752  0.60995499  0.610253519  0.610562309
#> carb        -3.21373036 -3.21353051 -3.213327276 -3.213120970
#> cyl         -3.59295041 -3.59263990 -3.592336464 -3.592040008
#> disp        -0.02030107 -0.02029783 -0.020294608 -0.020291403
#> drat         2.23921332  2.23914029  2.239060351  2.238973878
#> gear         0.47893694  0.47900288  0.479050202  0.479079760
#> hp          -0.00402954 -0.00403397 -0.004038096 -0.004041929
#> qsec        -1.58706625 -1.58693049 -1.586792057 -1.586651192
#> vs          -3.59219733 -3.59168742 -3.591178385 -3.590670604
#> wt           2.77020479  2.76982262  2.769430062  2.769027875
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1   1  60  64  17   1   8   8   8   8   8   8   8   8
#>  [37]   7  47   1   1   1   1   1   1  75   1  72   1   1   1   1   1   1  82
#>  [55]   1   1  38   1   1   1 132   1   1   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1 106   1   1   1   1   1   2  76   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.970626570 5.568226322 5.192946504 4.842959290 4.516560042 4.212159012
#>   [7] 3.928273592 3.663521099 3.416612039 3.186343822 2.971594913 2.771319361
#>  [13] 2.584541712 2.410352251 2.247902576 2.096401466 1.955111024 1.823343084
#>  [19] 1.700455863 1.585850829 1.478969791 1.379292177 1.286332500 1.199637995
#>  [25] 1.118786409 1.043383950 0.973063363 0.907482150 0.846320890 0.789281694
#>  [31] 0.736086750 0.686476967 0.640210717 0.597062657 0.556822632 0.519294650
#>  [37] 0.484295929 0.451656004 0.421215901 0.392827359 0.366352110 0.341661204
#>  [43] 0.318634382 0.297159491 0.277131936 0.258454171 0.241035225 0.224790258
#>  [49] 0.209640147 0.195511102 0.182334308 0.170045586 0.158585083 0.147896979
#>  [55] 0.137929218 0.128633250 0.119963799 0.111878641 0.104338395 0.097306336
#>  [61] 0.090748215 0.084632089 0.078928169 0.073608675 0.068647697 0.064021072
#>  [67] 0.059706266 0.055682263 0.051929465 0.048429593 0.045165600 0.042121590
#>  [73] 0.039282736 0.036635211 0.034166120 0.031863438 0.029715949 0.027713194
#>  [79] 0.025845417 0.024103523 0.022479026 0.020964015 0.019551110 0.018233431
#>  [85] 0.017004559 0.015858508 0.014789698 0.013792922 0.012863325 0.011996380
#>  [91] 0.011187864 0.010433839 0.009730634 0.009074821 0.008463209 0.007892817
#>  [97] 0.007360867 0.006864770 0.006402107 0.005970627
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 24
#> 
#> $loss
#>   [1] 933.33238 789.62074 669.38459 569.24561 486.28714 417.99257 362.19185
#>   [8] 317.01513 280.85245 252.31878 230.22372 213.54518 201.40656 193.05696
#>  [15] 187.85405 185.24920 184.71637 184.71637 184.71637 184.71637 184.71637
#>  [22] 184.71637 184.71637 184.71637 147.09850 114.45154 110.64849 110.64849
#>  [29] 106.86221 103.05812  99.94651  97.42204  95.39595  93.79191  92.54427
#>  [36]  91.59667  90.89970  74.52801  74.52801  74.52800  74.52800  74.52800
#>  [43]  74.52800  74.52800  65.82594  65.82593  61.91540  61.91538  61.91537
#>  [50]  61.91536  61.91535  61.91535  61.91534  57.75587  57.75586  57.75585
#>  [57]  56.72517  56.72517  56.72516  56.72516  50.90974  50.90972  50.90970
#>  [64]  50.90969  50.90967  50.90966  50.90965  50.90963  50.90962  50.90962
#>  [71]  50.90961  50.90960  50.90959  50.90959  50.90958  50.90957  50.90957
#>  [78]  50.90956  50.79069  50.79067  50.79065  50.79063  50.79061  50.79059
#>  [85]  50.79056  50.70795  50.70793  50.70791  50.70789  50.70786  50.70784
#>  [92]  50.70782  50.70780  50.70778  50.70777  50.70775  50.70773  50.70771
#>  [99]  50.70769  50.70767
#> 
#> $linear.predictors
#>      5.9706   5.5682   5.1929   4.8430   4.5166   4.2122   3.9283   3.6635
#> 1  19.71905 19.92814 20.12314 20.30500 20.47460 20.63277 20.78029 20.91785
#> 2  19.71905 19.92814 20.12314 20.30500 20.47460 20.63277 20.78029 20.91785
#> 3  19.71905 19.19631 18.70881 18.25416 17.83016 17.43473 17.06595 16.72203
#> 4  19.71905 19.92814 20.12314 20.30500 20.47460 20.63277 20.78029 20.91785
#> 5  19.71905 20.65997 21.53747 22.35584 23.11905 23.83082 24.49462 25.11368
#> 6  19.71905 19.92814 20.12314 20.30500 20.47460 20.63277 20.78029 20.91785
#> 7  19.71905 19.19631 18.70881 18.25416 17.83016 17.43473 17.06595 16.72203
#> 8  19.71905 19.19631 18.70881 18.25416 17.83016 17.43473 17.06595 16.72203
#> 9  19.71905 19.19631 18.70881 18.25416 17.83016 17.43473 17.06595 16.72203
#> 10 19.71905 19.19631 18.70881 18.25416 17.83016 17.43473 17.06595 16.72203
#> 11 19.71905 19.19631 18.70881 18.25416 17.83016 17.43473 17.06595 16.72203
#> 12 19.71905 20.65997 21.53747 22.35584 23.11905 23.83082 24.49462 25.11368
#> 13 19.71905 20.65997 21.53747 22.35584 23.11905 23.83082 24.49462 25.11368
#> 14 19.71905 20.65997 21.53747 22.35584 23.11905 23.83082 24.49462 25.11368
#> 15 19.71905 19.19631 18.70881 18.25416 17.83016 17.43473 17.06595 16.72203
#> 16 19.71905 19.19631 18.70881 18.25416 17.83016 17.43473 17.06595 16.72203
#> 17 19.71905 19.19631 18.70881 18.25416 17.83016 17.43473 17.06595 16.72203
#> 18 19.71905 19.19631 18.70881 18.25416 17.83016 17.43473 17.06595 16.72203
#> 19 19.71905 20.65997 21.53747 22.35584 23.11905 23.83082 24.49462 25.11368
#> 20 19.71905 19.19631 18.70881 18.25416 17.83016 17.43473 17.06595 16.72203
#> 21 19.71905 19.92814 20.12314 20.30500 20.47460 20.63277 20.78029 20.91785
#>      3.4166   3.1863   2.9716   2.7713   2.5845   2.4104   2.2479   2.0964
#> 1  21.04615 21.16580 21.27739 21.38146 21.47851 21.56902 21.65343 21.73215
#> 2  21.04615 21.16580 21.27739 21.38146 21.47851 21.56902 21.65343 21.73215
#> 3  16.40129 16.10216 15.82319 15.56303 15.32039 15.09412 14.88309 14.68628
#> 4  21.04615 21.16580 21.27739 21.38146 21.47851 21.56902 21.65343 21.73215
#> 5  25.69102 26.22945 26.73159 27.19989 27.63662 28.04393 28.42378 28.77803
#> 6  21.04615 21.16580 21.27739 21.38146 21.47851 21.56902 21.65343 21.73215
#> 7  16.40129 16.10216 15.82319 15.56303 15.32039 15.09412 14.88309 14.68628
#> 8  16.40129 16.10216 15.82319 15.56303 15.32039 15.09412 14.88309 14.68628
#> 9  16.40129 16.10216 15.82319 15.56303 15.32039 15.09412 14.88309 14.68628
#> 10 16.40129 16.10216 15.82319 15.56303 15.32039 15.09412 14.88309 14.68628
#> 11 16.40129 16.10216 15.82319 15.56303 15.32039 15.09412 14.88309 14.68628
#> 12 25.69102 26.22945 26.73159 27.19989 27.63662 28.04393 28.42378 28.77803
#> 13 25.69102 26.22945 26.73159 27.19989 27.63662 28.04393 28.42378 28.77803
#> 14 25.69102 26.22945 26.73159 27.19989 27.63662 28.04393 28.42378 28.77803
#> 15 16.40129 16.10216 15.82319 15.56303 15.32039 15.09412 14.88309 14.68628
#> 16 16.40129 16.10216 15.82319 15.56303 15.32039 15.09412 14.88309 14.68628
#> 17 16.40129 16.10216 15.82319 15.56303 15.32039 15.09412 14.88309 14.68628
#> 18 16.40129 16.10216 15.82319 15.56303 15.32039 15.09412 14.88309 14.68628
#> 19 25.69102 26.22945 26.73159 27.19989 27.63662 28.04393 28.42378 28.77803
#> 20 16.40129 16.10216 15.82319 15.56303 15.32039 15.09412 14.88309 14.68628
#> 21 21.04615 21.16580 21.27739 21.38146 21.47851 21.56902 21.65343 21.73215
#>      1.9551   1.8233   1.7005   1.5859   1.4790   1.3793   1.2863   1.1996
#> 1  21.78733 21.78733 21.78733 21.78733 21.78733 21.78733 21.78733 21.78733
#> 2  21.78733 21.78733 21.78733 21.78733 21.78733 21.78733 21.78733 21.78733
#> 3  14.54833 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833
#> 4  21.78733 21.78733 21.78733 21.78733 21.78733 21.78733 21.78733 21.78733
#> 5  29.02633 29.02633 29.02633 29.02633 29.02633 29.02633 29.02633 29.02633
#> 6  21.78733 21.78733 21.78733 21.78733 21.78733 21.78733 21.78733 21.78733
#> 7  14.54833 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833
#> 8  14.54833 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833
#> 9  14.54833 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833
#> 10 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833
#> 11 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833
#> 12 29.02633 29.02633 29.02633 29.02633 29.02633 29.02633 29.02633 29.02633
#> 13 29.02633 29.02633 29.02633 29.02633 29.02633 29.02633 29.02633 29.02633
#> 14 29.02633 29.02633 29.02633 29.02633 29.02633 29.02633 29.02633 29.02633
#> 15 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833
#> 16 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833
#> 17 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833
#> 18 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833
#> 19 29.02633 29.02633 29.02633 29.02633 29.02633 29.02633 29.02633 29.02633
#> 20 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833 14.54833
#> 21 21.78733 21.78733 21.78733 21.78733 21.78733 21.78733 21.78733 21.78733
#>      1.1188   1.0434   0.9731   0.9075   0.8463   0.7893   0.7361   0.6865
#> 1  22.18116 22.80748 23.10593 23.10601 22.98315 22.84464 22.71551 22.59509
#> 2  21.67981 21.50847 21.42675 21.42667 21.55313 21.69571 21.82864 21.95261
#> 3  15.10811 15.99943 16.42436 16.42467 16.50728 16.59909 16.68453 16.76420
#> 4  21.47338 20.97358 20.73533 20.73518 20.86712 21.01610 21.15503 21.28460
#> 5  28.11670 26.66819 25.97760 25.97708 25.99362 26.01414 26.03349 26.05154
#> 6  21.49023 21.01725 20.79177 20.79162 20.68709 20.57003 20.46099 20.35930
#> 7  14.57728 14.62401 14.64641 14.64654 14.66456 14.68434 14.70272 14.71985
#> 8  14.82163 15.25714 15.46483 15.46505 15.47658 15.48878 15.50005 15.51056
#> 9  13.58302 12.04782 11.31627 11.31609 11.28181 11.24394 11.20873 11.17589
#> 10 13.43640 11.66794 10.82522 10.82498 10.79460 10.76128 10.73033 10.70147
#> 11 13.50297 11.84042 11.04817 11.04796 11.01580 10.98042 10.94753 10.91687
#> 12 28.91716 28.74224 28.65864 28.65838 28.73236 28.81654 28.89512 28.96840
#> 13 29.41008 30.01942 30.30960 30.30950 30.29173 30.27210 30.25385 30.23683
#> 14 29.22471 29.53911 29.68873 29.68857 29.75439 29.82902 29.89865 29.96359
#> 15 15.04070 15.82478 16.19859 16.19888 16.28328 16.37717 16.46457 16.54608
#> 16 15.11232 16.01035 16.43847 16.43878 16.52128 16.61296 16.69827 16.77784
#> 17 14.77107 15.12615 15.29550 15.29570 15.22990 15.15516 15.08541 15.02036
#> 18 14.76686 15.11523 15.28139 15.28159 15.37325 15.47565 15.57101 15.65994
#> 19 29.49603 30.24210 30.59746 30.59739 30.57733 30.55504 30.53429 30.51494
#> 20 15.33561 16.58890 17.18634 17.18672 17.10595 17.01368 16.92752 16.84716
#> 21 22.05477 22.48000 22.68261 22.68265 22.40578 22.09419 21.80375 21.53289
#>      0.6402   0.5971   0.5568   0.5193   0.4843   0.4517   0.4212   0.3928
#> 1  22.48278 22.37804 22.28036 22.18927 22.10423 22.46144 22.46129 22.46121
#> 2  22.06822 22.17604 22.27660 22.37038 22.45791 22.11541 22.11545 22.11547
#> 3  16.83851 16.90781 16.97244 17.03271 17.08932 16.23921 16.23908 16.23900
#> 4  21.40543 21.51812 21.62322 21.72123 21.81265 22.11541 22.11545 22.11547
#> 5  26.06838 26.08407 26.09871 26.11237 26.12461 24.49389 24.49440 24.49469
#> 6  20.26446 20.17602 20.09353 20.01661 19.94460 16.86883 16.86932 16.86960
#> 7  14.73584 14.75074 14.76464 14.77760 14.78985 14.49035 14.49037 14.49038
#> 8  15.52036 15.52951 15.53803 15.54598 15.55362 14.49035 14.49037 14.49038
#> 9  11.14526 11.11670 11.09007 11.06523 11.04183 12.74149 12.74166 12.74176
#> 10 10.67455 10.64944 10.62603 10.60420 10.58357 12.74149 12.74166 12.74176
#> 11 10.88826 10.86159 10.83671 10.81352 10.79163 12.74149 12.74166 12.74176
#> 12 29.03674 29.10047 29.15991 29.21534 29.26688 31.83536 31.83508 31.83492
#> 13 30.22096 30.20616 30.19235 30.17948 30.16735 30.08650 30.08637 30.08630
#> 14 30.02416 30.08064 30.13331 30.18244 30.22818 31.83536 31.83508 31.83492
#> 15 16.62209 16.69298 16.75909 16.82075 16.87863 16.23921 16.23908 16.23900
#> 16 16.85204 16.92124 16.98577 17.04596 17.10249 16.23921 16.23908 16.23900
#> 17 14.95969 14.90311 14.85035 14.80114 14.75536 12.74149 12.74166 12.74176
#> 18 15.74288 15.82022 15.89236 15.95963 16.02267 16.23921 16.23908 16.23900
#> 19 30.49690 30.48007 30.46437 30.44973 30.43599 30.08650 30.08637 30.08630
#> 20 16.77222 16.70233 16.63715 16.57636 16.51994 18.33410 18.33363 18.33336
#> 21 21.28028 21.04469 20.82499 20.62009 20.42869 18.96372 18.96388 18.96397
#>      0.3664   0.3417   0.3186   0.2972   0.2771   0.2585   0.2410   0.2248
#> 1  22.46116 22.46113 22.46112 22.46111 22.19358 22.19340 22.00016 22.00034
#> 2  22.11548 22.11549 22.11549 22.11550 22.46259 22.46277 22.19625 22.19609
#> 3  16.23896 16.23893 16.23892 16.23891 16.60556 16.60560 16.73524 16.73519
#> 4  22.11548 22.11549 22.11549 22.11550 21.82522 21.82521 21.17335 21.17337
#> 5  24.49486 24.49496 24.49502 24.49505 23.60038 23.60044 23.96944 23.96936
#> 6  16.86977 16.86986 16.86992 16.86995 17.59002 17.59045 18.09925 18.09881
#> 7  14.49038 14.49039 14.49039 14.49039 14.36033 14.36028 14.31071 14.31076
#> 8  14.49038 14.49039 14.49039 14.49039 13.87004 13.86985 13.84138 13.84157
#> 9  12.74181 12.74184 12.74186 12.74187 11.95167 11.95149 11.65234 11.65253
#> 10 12.74181 12.74184 12.74186 12.74187 12.08242 12.08227 11.86779 11.86794
#> 11 12.74181 12.74184 12.74186 12.74187 12.40928 12.40922 12.47736 12.47741
#> 12 31.83483 31.83478 31.83475 31.83473 31.43792 31.43774 31.37034 31.37052
#> 13 30.08626 30.08624 30.08622 30.08621 30.27951 30.27954 31.18580 31.18574
#> 14 31.83483 31.83478 31.83475 31.83473 31.08655 31.08627 31.21458 31.21486
#> 15 16.23896 16.23893 16.23892 16.23891 16.72813 16.72820 16.34949 16.34943
#> 16 16.23896 16.23893 16.23892 16.23891 16.37675 16.37673 16.51622 16.51624
#> 17 12.74181 12.74184 12.74186 12.74187 14.05175 14.05216 14.69458 14.69415
#> 18 16.23896 16.23893 16.23892 16.23891 16.58104 16.58107 16.62147 16.62144
#> 19 30.08626 30.08624 30.08622 30.08621 31.60329 31.60370 30.95664 30.95626
#> 20 18.33321 18.33312 18.33307 18.33304 17.89536 17.89495 18.52590 18.52630
#> 21 18.96402 18.96405 18.96407 18.96408 19.10863 19.10867 18.34171 18.34170
#>      0.2096   0.1955   0.1823   0.1700   0.1586   0.1479   0.1379   0.1286
#> 1  22.00051 22.00065 22.00078 22.00089 22.00099 22.59325 22.59331 22.59338
#> 2  22.19595 22.19582 22.19571 22.19561 22.19552 21.57878 21.57868 21.57860
#> 3  16.73515 16.73512 16.73508 16.73506 16.73503 17.13092 17.13103 17.13113
#> 4  21.17340 21.17342 21.17343 21.17345 21.17347 20.40046 20.40025 20.40004
#> 5  23.96930 23.96924 23.96919 23.96914 23.96910 24.68531 24.68560 24.68587
#> 6  18.09842 18.09807 18.09777 18.09751 18.09727 17.17380 17.17377 17.17375
#> 7  14.31080 14.31084 14.31087 14.31090 14.31092 14.44099 14.44100 14.44100
#> 8  13.84173 13.84187 13.84200 13.84211 13.84220 13.90991 13.90984 13.90978
#> 9  11.65270 11.65285 11.65298 11.65309 11.65319 11.48256 11.48243 11.48231
#> 10 11.86808 11.86819 11.86830 11.86839 11.86847 11.73091 11.73080 11.73071
#> 11 12.47746 12.47750 12.47753 12.47756 12.47759 12.43565 12.43561 12.43558
#> 12 31.37068 31.37082 31.37095 31.37106 31.37115 31.36324 31.36317 31.36311
#> 13 31.18568 31.18563 31.18559 31.18555 31.18551 31.26852 31.26860 31.26868
#> 14 31.21510 31.21531 31.21550 31.21566 31.21580 31.19608 31.19598 31.19589
#> 15 16.34937 16.34932 16.34928 16.34924 16.34921 16.66906 16.66914 16.66922
#> 16 16.51625 16.51627 16.51628 16.51629 16.51629 16.88308 16.88316 16.88323
#> 17 14.69378 14.69345 14.69316 14.69290 14.69268 14.97713 14.97738 14.97762
#> 18 16.62141 16.62138 16.62135 16.62133 16.62131 16.99763 16.99773 16.99782
#> 19 30.95592 30.95563 30.95537 30.95515 30.95495 30.93380 30.93394 30.93407
#> 20 18.52664 18.52695 18.52721 18.52745 18.52765 17.69605 17.69564 17.69526
#> 21 18.34169 18.34168 18.34168 18.34168 18.34167 18.55288 18.55292 18.55296
#>      0.1200   0.1119   0.1043   0.0973   0.0907   0.0846   0.0789   0.0736
#> 1  22.71476 22.71506 22.71529 22.71546 22.90906 22.90906 22.90907 22.90907
#> 2  21.43178 21.43149 21.43126 21.43108 21.07012 21.07003 21.06995 21.06987
#> 3  17.23506 17.23502 17.23499 17.23497 16.44610 16.44594 16.44578 16.44563
#> 4  20.30365 20.30344 20.30327 20.30314 20.46689 20.46691 20.46692 20.46693
#> 5  24.87882 24.87939 24.87983 24.88018 23.94454 23.94435 23.94416 23.94398
#> 6  17.75127 17.75093 17.75068 17.75049 17.71410 17.71429 17.71448 17.71466
#> 7  14.95130 14.95132 14.95134 14.95137 15.80835 15.80865 15.80893 15.80920
#> 8  14.35929 14.35939 14.35947 14.35954 14.23731 14.23734 14.23738 14.23741
#> 9  10.77421 10.77424 10.77425 10.77426 10.65791 10.65771 10.65751 10.65732
#> 10 11.11973 11.11975 11.11976 11.11976 11.68171 11.68170 11.68169 11.68168
#> 11 12.02055 12.02055 12.02054 12.02054 12.94845 12.94860 12.94874 12.94887
#> 12 31.13997 31.14012 31.14024 31.14033 32.36074 32.36099 32.36124 32.36147
#> 13 31.29400 31.29412 31.29422 31.29430 31.18730 31.18736 31.18742 31.18747
#> 14 30.99778 30.99801 30.99819 30.99833 31.30072 31.30077 31.30081 31.30085
#> 15 16.95625 16.95616 16.95609 16.95604 16.85312 16.85312 16.85312 16.85312
#> 16 17.27705 17.27707 17.27708 17.27710 17.09797 17.09801 17.09804 17.09807
#> 17 15.36852 15.36833 15.36818 15.36807 15.12916 15.12924 15.12931 15.12938
#> 18 16.85867 16.85862 16.85858 16.85855 16.44416 16.44403 16.44391 16.44379
#> 19 30.80272 30.80250 30.80231 30.80215 30.55154 30.55143 30.55132 30.55122
#> 20 17.09687 17.09662 17.09642 17.09628 16.94157 16.94148 16.94140 16.94131
#> 21 18.76775 18.76789 18.76799 18.76806 18.34916 18.34899 18.34883 18.34867
#>      0.0686   0.0640   0.0597   0.0557   0.0519   0.0484   0.0452   0.0421
#> 1  22.90908 22.90908 22.90908 22.90909 22.90909 22.90909 22.90910 22.90910
#> 2  21.06979 21.06972 21.06965 21.06958 21.06951 21.06945 21.06939 21.06934
#> 3  16.44548 16.44534 16.44520 16.44508 16.44495 16.44483 16.44472 16.44461
#> 4  20.46694 20.46695 20.46696 20.46697 20.46698 20.46699 20.46700 20.46701
#> 5  23.94381 23.94365 23.94349 23.94334 23.94319 23.94305 23.94292 23.94279
#> 6  17.71483 17.71500 17.71516 17.71531 17.71545 17.71559 17.71573 17.71586
#> 7  15.80946 15.80971 15.80995 15.81018 15.81040 15.81061 15.81081 15.81100
#> 8  14.23745 14.23748 14.23751 14.23754 14.23757 14.23759 14.23762 14.23764
#> 9  10.65714 10.65697 10.65681 10.65665 10.65650 10.65635 10.65622 10.65608
#> 10 11.68167 11.68166 11.68166 11.68165 11.68164 11.68164 11.68163 11.68162
#> 11 12.94900 12.94912 12.94924 12.94935 12.94946 12.94956 12.94966 12.94975
#> 12 32.36170 32.36191 32.36212 32.36232 32.36250 32.36269 32.36286 32.36302
#> 13 31.18753 31.18758 31.18763 31.18767 31.18772 31.18776 31.18780 31.18784
#> 14 31.30089 31.30093 31.30096 31.30100 31.30103 31.30106 31.30109 31.30112
#> 15 16.85312 16.85312 16.85312 16.85311 16.85311 16.85311 16.85311 16.85311
#> 16 17.09811 17.09814 17.09817 17.09820 17.09822 17.09825 17.09827 17.09830
#> 17 15.12945 15.12951 15.12958 15.12964 15.12969 15.12974 15.12980 15.12984
#> 18 16.44368 16.44358 16.44347 16.44337 16.44328 16.44319 16.44310 16.44302
#> 19 30.55112 30.55103 30.55093 30.55085 30.55076 30.55068 30.55060 30.55053
#> 20 16.94124 16.94116 16.94109 16.94102 16.94096 16.94090 16.94084 16.94079
#> 21 18.34852 18.34838 18.34824 18.34811 18.34798 18.34786 18.34774 18.34763
#>      0.0393   0.0366   0.0342   0.0319   0.0297   0.0277   0.0258   0.0241
#> 1  22.90910 22.90910 22.90910 22.90911 22.90911 22.90911 22.81886 22.81876
#> 2  21.06928 21.06923 21.06918 21.06914 21.06909 21.06905 20.98306 20.98290
#> 3  16.44451 16.44441 16.44431 16.44422 16.44413 16.44405 16.42847 16.42838
#> 4  20.46701 20.46702 20.46703 20.46704 20.46705 20.46706 20.41711 20.41703
#> 5  23.94267 23.94255 23.94244 23.94233 23.94222 23.94212 24.07433 24.07440
#> 6  17.71598 17.71610 17.71621 17.71632 17.71642 17.71652 17.74857 17.74869
#> 7  15.81119 15.81136 15.81153 15.81169 15.81185 15.81200 15.87171 15.87194
#> 8  14.23767 14.23769 14.23772 14.23774 14.23776 14.23778 14.20271 14.20267
#> 9  10.65596 10.65583 10.65572 10.65561 10.65550 10.65540 10.58236 10.58216
#> 10 11.68162 11.68161 11.68160 11.68160 11.68159 11.68159 11.67382 11.67381
#> 11 12.94984 12.94993 12.95001 12.95009 12.95016 12.95023 12.98077 12.98090
#> 12 32.36318 32.36334 32.36348 32.36362 32.36375 32.36388 32.40386 32.40405
#> 13 31.18788 31.18791 31.18795 31.18798 31.18801 31.18804 31.05594 31.05583
#> 14 31.30115 31.30118 31.30120 31.30123 31.30125 31.30128 31.25492 31.25488
#> 15 16.85311 16.85310 16.85310 16.85310 16.85310 16.85310 16.88670 16.88675
#> 16 17.09832 17.09834 17.09836 17.09838 17.09840 17.09842 17.13211 17.13217
#> 17 15.12989 15.12993 15.12998 15.13002 15.13005 15.13009 15.04865 15.04863
#> 18 16.44294 16.44286 16.44279 16.44272 16.44265 16.44258 16.46770 16.46768
#> 19 30.55045 30.55039 30.55032 30.55025 30.55019 30.55013 30.65814 30.65823
#> 20 16.94073 16.94069 16.94064 16.94059 16.94055 16.94051 17.07559 17.07563
#> 21 18.34752 18.34742 18.34732 18.34723 18.34714 18.34705 18.33463 18.33451
#>      0.0225   0.0210   0.0196   0.0182   0.0170   0.0159   0.0148   0.0138
#> 1  22.81867 22.81858 22.81850 22.81841 22.81819 22.89224 22.89229 22.89235
#> 2  20.98276 20.98261 20.98247 20.98233 20.98208 20.94477 20.94490 20.94503
#> 3  16.42828 16.42819 16.42810 16.42801 16.42776 16.42607 16.42615 16.42624
#> 4  20.41696 20.41689 20.41682 20.41675 20.41665 20.35969 20.35976 20.35983
#> 5  24.07447 24.07453 24.07460 24.07466 24.07504 24.04779 24.04769 24.04760
#> 6  17.74881 17.74893 17.74905 17.74916 17.74962 17.87755 17.87790 17.87823
#> 7  15.87216 15.87238 15.87259 15.87280 15.87317 15.86766 15.86746 15.86726
#> 8  14.20262 14.20258 14.20254 14.20250 14.20244 14.11037 14.11028 14.11020
#> 9  10.58196 10.58177 10.58159 10.58141 10.58107 10.60090 10.60107 10.60124
#> 10 11.67380 11.67379 11.67378 11.67377 11.67375 11.69885 11.69885 11.69884
#> 11 12.98102 12.98115 12.98127 12.98139 12.98160 12.97753 12.97732 12.97711
#> 12 32.40424 32.40442 32.40460 32.40477 32.40504 32.42000 32.41981 32.41962
#> 13 31.05572 31.05561 31.05551 31.05541 31.05526 31.11073 31.11086 31.11099
#> 14 31.25484 31.25480 31.25476 31.25472 31.25463 31.19675 31.19667 31.19659
#> 15 16.88680 16.88685 16.88689 16.88694 16.88691 16.96591 16.96603 16.96613
#> 16 17.13223 17.13229 17.13235 17.13240 17.13244 17.20530 17.20543 17.20556
#> 17 15.04862 15.04860 15.04859 15.04858 15.04855 14.88667 14.88619 14.88573
#> 18 16.46766 16.46764 16.46762 16.46760 16.46746 16.53142 16.53155 16.53168
#> 19 30.65832 30.65840 30.65849 30.65857 30.65867 30.64393 30.64378 30.64362
#> 20 17.07566 17.07569 17.07572 17.07576 17.07574 17.03652 17.03673 17.03693
#> 21 18.33440 18.33429 18.33418 18.33408 18.33393 18.29934 18.29927 18.29922
#>      0.0129   0.0120   0.0112   0.0104   0.0097   0.0091   0.0085   0.0079
#> 1  22.89240 22.89245 22.89250 22.89255 22.89259 22.89264 22.89268 22.89272
#> 2  20.94516 20.94530 20.94544 20.94558 20.94572 20.94586 20.94601 20.94615
#> 3  16.42632 16.42640 16.42648 16.42657 16.42665 16.42673 16.42681 16.42690
#> 4  20.35991 20.35999 20.36007 20.36016 20.36025 20.36034 20.36043 20.36053
#> 5  24.04750 24.04740 24.04731 24.04722 24.04712 24.04703 24.04694 24.04685
#> 6  17.87853 17.87882 17.87909 17.87935 17.87958 17.87980 17.88001 17.88020
#> 7  15.86706 15.86686 15.86665 15.86645 15.86625 15.86605 15.86585 15.86566
#> 8  14.11013 14.11007 14.11002 14.10998 14.10994 14.10991 14.10989 14.10987
#> 9  10.60141 10.60159 10.60176 10.60193 10.60210 10.60227 10.60244 10.60261
#> 10 11.69884 11.69884 11.69883 11.69882 11.69882 11.69881 11.69880 11.69879
#> 11 12.97691 12.97671 12.97651 12.97631 12.97611 12.97592 12.97573 12.97554
#> 12 32.41944 32.41925 32.41906 32.41887 32.41868 32.41849 32.41831 32.41812
#> 13 31.11111 31.11123 31.11135 31.11146 31.11157 31.11168 31.11179 31.11189
#> 14 31.19652 31.19645 31.19640 31.19635 31.19630 31.19626 31.19623 31.19620
#> 15 16.96623 16.96632 16.96641 16.96648 16.96655 16.96661 16.96666 16.96671
#> 16 17.20568 17.20579 17.20590 17.20599 17.20608 17.20616 17.20623 17.20629
#> 17 14.88529 14.88487 14.88447 14.88409 14.88373 14.88339 14.88306 14.88275
#> 18 16.53179 16.53190 16.53201 16.53211 16.53220 16.53229 16.53237 16.53244
#> 19 30.64347 30.64331 30.64316 30.64301 30.64286 30.64272 30.64257 30.64243
#> 20 17.03713 17.03732 17.03751 17.03769 17.03787 17.03805 17.03822 17.03839
#> 21 18.29916 18.29912 18.29908 18.29905 18.29902 18.29900 18.29898 18.29897
#>      0.0074   0.0069   0.0064   0.0060
#> 1  22.89276 22.89279 22.89283 22.89286
#> 2  20.94630 20.94645 20.94660 20.94675
#> 3  16.42698 16.42706 16.42714 16.42722
#> 4  20.36062 20.36072 20.36082 20.36093
#> 5  24.04676 24.04668 24.04659 24.04651
#> 6  17.88037 17.88053 17.88068 17.88081
#> 7  15.86546 15.86526 15.86507 15.86487
#> 8  14.10986 14.10985 14.10986 14.10986
#> 9  10.60277 10.60294 10.60310 10.60327
#> 10 11.69878 11.69877 11.69876 11.69875
#> 11 12.97536 12.97518 12.97500 12.97482
#> 12 32.41793 32.41775 32.41756 32.41737
#> 13 31.11199 31.11209 31.11219 31.11228
#> 14 31.19618 31.19616 31.19614 31.19613
#> 15 16.96675 16.96678 16.96681 16.96683
#> 16 17.20635 17.20640 17.20644 17.20648
#> 17 14.88246 14.88219 14.88193 14.88169
#> 18 16.53251 16.53258 16.53264 16.53270
#> 19 30.64229 30.64215 30.64201 30.64187
#> 20 17.03855 17.03871 17.03887 17.03902
#> 21 18.29896 18.29896 18.29897 18.29897
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.4 18.7 18.1 22.8 19.2 16.4 15.2 10.4 10.4 14.7 32.4 30.4 33.9 15.5
#> [16] 15.2 13.3 19.2 30.4 15.8 19.7
#> 
#> $X
#>               am       carb        cyl         disp       drat       gear
#>  [1,]  1.4142136  0.9227055 -0.3464102 -0.765587399  0.6963610  0.5883484
#>  [2,] -0.7071068 -1.3130809 -0.3464102 -0.002077578 -0.7466398 -0.7844645
#>  [3,] -0.7071068 -0.5678188  0.8660254  0.792595909 -0.6234568 -0.7844645
#>  [4,] -0.7071068 -1.3130809 -0.3464102 -0.259177824 -1.3097620 -0.7844645
#>  [5,] -0.7071068 -0.5678188 -1.5588457 -0.915172996  0.7315561  0.5883484
#>  [6,] -0.7071068  0.9227055 -0.3464102 -0.706376433  0.7315561  0.5883484
#>  [7,] -0.7071068  0.1774434  0.8660254  0.136600737 -0.7642373 -0.7844645
#>  [8,] -0.7071068  0.1774434  0.8660254  0.136600737 -0.7642373 -0.7844645
#>  [9,] -0.7071068  0.9227055  0.8660254  1.665178562 -1.0106033 -0.7844645
#> [10,] -0.7071068  0.9227055  0.8660254  1.571687563 -0.8874203 -0.7844645
#> [11,] -0.7071068  0.9227055  0.8660254  1.415869233 -0.4826762 -0.7844645
#> [12,]  1.4142136 -1.3130809 -1.5588457 -1.398988913  1.0131173  0.5883484
#> [13,]  1.4142136 -0.5678188 -1.5588457 -1.422361663  2.5089108  0.5883484
#> [14,]  1.4142136 -1.3130809 -1.5588457 -1.458199879  1.2594833  0.5883484
#> [15,] -0.7071068 -0.5678188  0.8660254  0.465377415 -1.3097620 -0.7844645
#> [16,] -0.7071068 -0.5678188  0.8660254  0.356304583 -0.6234568 -0.7844645
#> [17,] -0.7071068  0.9227055  0.8660254  0.714686744  0.3972023 -0.7844645
#> [18,] -0.7071068 -0.5678188  0.8660254  1.104232571 -0.7466398 -0.7844645
#> [19,]  1.4142136 -0.5678188 -1.5588457 -1.271217882  0.4675926  1.9611614
#> [20,]  1.4142136  0.9227055  0.8660254  0.722477660  1.2594833  1.9611614
#> [21,]  1.4142136  2.4132297 -0.3464102 -0.882451147  0.2036290  1.9611614
#>                hp         qsec         vs          wt
#>  [1,] -0.70255568 -0.750858629 -0.7844645 -0.70766820
#>  [2,] -0.70255568  0.885239317  1.2747549 -0.15214177
#>  [3,]  0.39594203 -0.443403981 -0.7844645  0.05793125
#>  [4,] -0.78705551  1.313479718  1.2747549  0.07660440
#>  [5,] -0.95605516  2.784869817  1.2747549 -0.21282953
#>  [6,] -0.48285614  0.259349499  1.2747549  0.05793125
#>  [7,]  0.48044186 -0.234774042 -0.7844645  0.64613570
#>  [8,]  0.48044186  0.094641652 -0.7844645  0.37537492
#>  [9,]  0.90294098  0.083661129 -0.7844645  1.74785198
#> [10,]  1.07194063 -0.004183056 -0.7844645  1.91030845
#> [11,]  1.32544010 -0.223793519 -0.7844645  1.83654948
#> [12,] -1.44615414  0.901710101  1.2747549 -1.09980450
#> [13,] -1.68275365  0.380135253  1.2747549 -1.64599435
#> [14,] -1.46305410  1.137791348  1.2747549 -1.44058962
#> [15,] -0.02655709 -0.525757905 -0.7844645  0.13262388
#> [16,] -0.02655709 -0.289676658 -0.7844645  0.05326296
#> [17,]  1.57893958 -1.327336093 -0.7844645  0.43139439
#> [18,]  0.39594203 -0.426933197 -0.7844645  0.43606268
#> [19,] -0.65185579 -0.509287120  1.2747549 -1.74122746
#> [20,]  1.90003891 -1.826949895 -0.7844645 -0.19415638
#> [21,]  0.39594203 -1.277923739 -0.7844645 -0.56761952
#> attr(,"center")
#>  [1]   0.3333333   2.7619048   6.5714286 258.2666667   3.5042857   3.5714286
#>  [7] 151.5714286  17.8276190   0.3809524   3.3779524
#> attr(,"scale")
#>  [1]   0.4714045   1.3418098   1.6495722 128.3546031   0.5682603   0.7284314
#>  [7]  59.1717206   1.8214068   0.4856209   1.0710562
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 36
#> 
#> $lambda.min
#> [1] 0.5192947
#> 
#> $null.dev
#> [1] 44.4444
#> 
#> $Bias
#> [1] 4.730354
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 13.50923