Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 35.851084 34.398982 31.136460 27.546631 24.533013 22.013233 19.916140
#>   [8] 18.180309 16.752749 15.587768 14.645999 13.893552 13.301276 12.844118
#>  [15] 12.500572 12.247305 11.943717 11.733805 11.600857 11.289058 10.916685
#>  [22] 10.458538  9.874156  9.174925  8.327268  7.492661  7.119375  6.935704
#>  [29]  6.305736  6.331663  6.356074  6.354940  6.365395  6.373093  6.381118
#>  [36]  6.382005  6.349757  6.314869  6.938059  6.912958  6.875376  6.997900
#>  [43]  7.804992  7.868667  8.102400  8.015800  7.978966  8.336661 10.801559
#>  [50] 12.971139 14.109862 14.189918 14.664725 14.661244 14.660657 14.697426
#>  [57] 15.928105 15.765841 15.765645 15.765464 16.015935 16.701733 16.356599
#>  [64] 16.369855 16.498350 18.129858 18.130499 18.776495 18.777313 19.126196
#>  [71] 19.157998 19.459243 19.460343 19.533682 20.524559 21.064066 21.064732
#>  [78] 21.014353 21.014940 21.015510 21.016064 21.242756 21.202966 21.303466
#>  [85] 21.303644 21.316259 21.316454 21.357999 21.311162 21.310947 21.310764
#>  [92] 21.361989 21.361869 21.361488 21.339651 21.339400 21.339419 21.371616
#>  [99] 21.371502 21.371378
#> 
#> $cvse
#>   [1] 11.655853 11.613616 10.994559  9.823334  8.791083  7.879381  7.072924
#>   [8]  6.359120  5.727722  5.170512  4.681001  4.254151  3.886089  3.573790
#>  [15]  3.314722  3.106151  2.938261  2.810276  2.728129  2.713185  2.656448
#>  [22]  2.529388  2.443596  2.449912  2.452879  2.436846  2.412075  2.402455
#>  [29]  2.417541  2.414372  2.413229  2.414549  2.415130  2.415758  2.415840
#>  [36]  2.414751  2.397521  2.382557  2.666340  2.655310  2.645165  2.620173
#>  [43]  3.153574  3.136297  3.150610  3.151582  3.148626  3.252244  3.896553
#>  [50]  4.594483  4.525167  4.496634  4.754646  4.784265  4.784102  4.840719
#>  [57]  5.013468  5.025740  5.025719  5.025697  5.121139  5.335463  5.115481
#>  [64]  5.113596  5.096949  5.126225  5.126321  5.071614  5.071679  5.236251
#>  [71]  5.237605  5.312266  5.312397  5.311798  5.691846  5.741030  5.741257
#>  [78]  5.746801  5.747021  5.747237  5.747447  5.740128  5.714697  5.757482
#>  [85]  5.757577  5.756042  5.756120  5.750257  5.754509  5.754599  5.754688
#>  [92]  5.762967  5.763071  5.763078  5.746373  5.746404  5.746453  5.751615
#>  [99]  5.751648  5.751678
#> 
#> $fold
#>  [1]  9  5  7  2  6  6  4  4  1  3  1  7  9  2  1 10  3  8  5 10  8
#> 
#> $lambda
#>   [1] 4.946072020 4.612723320 4.301841207 4.011911507 3.741522099 3.489356032
#>   [7] 3.254185114 3.034863928 2.830324257 2.639569875 2.461671700 2.295763266
#>  [13] 2.141036506 1.996737811 1.862164365 1.736660719 1.619615599 1.510458929
#>  [19] 1.408659052 1.313720147 1.225179806 1.142606787 1.065598914 0.993781114
#>  [25] 0.926803593 0.864340132 0.806086500 0.751758968 0.701092930 0.653841613
#>  [31] 0.609774876 0.568678091 0.530351091 0.494607202 0.461272332 0.430184121
#>  [37] 0.401191151 0.374152210 0.348935603 0.325418511 0.303486393 0.283032426
#>  [43] 0.263956987 0.246167170 0.229576327 0.214103651 0.199673781 0.186216437
#>  [49] 0.173666072 0.161961560 0.151045893 0.140865905 0.131372015 0.122517981
#>  [55] 0.114260679 0.106559891 0.099378111 0.092680359 0.086434013 0.080608650
#>  [61] 0.075175897 0.070109293 0.065384161 0.060977488 0.056867809 0.053035109
#>  [67] 0.049460720 0.046127233 0.043018412 0.040119115 0.037415221 0.034893560
#>  [73] 0.032541851 0.030348639 0.028303243 0.026395699 0.024616717 0.022957633
#>  [79] 0.021410365 0.019967378 0.018621644 0.017366607 0.016196156 0.015104589
#>  [85] 0.014086591 0.013137201 0.012251798 0.011426068 0.010655989 0.009937811
#>  [91] 0.009268036 0.008643401 0.008060865 0.007517590 0.007010929 0.006538416
#>  [97] 0.006097749 0.005686781 0.005303511 0.004946072
#> 
#> $fit
#> $beta
#>               4.9461     4.6127     4.3018     4.0119    3.7415    3.4894
#> (Intercept) 18.88571 20.8823551 22.7444290 24.4810053 26.100542 27.610928
#> am           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> cyl          0.00000 -0.3083048 -0.5958309 -0.8639788 -1.114054 -1.347276
#> disp         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> wt           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#>                3.2542    3.0349    2.8303    2.6396    2.4617    2.2958
#> (Intercept) 29.019518 30.333175 31.558295 32.700846 33.766393 34.760126
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.564779 -1.767623 -1.956796 -2.133219 -2.297752 -2.451196
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                2.1410    1.9967    1.8622    1.7367    1.6196    1.5105
#> (Intercept) 35.686885 36.551183 37.357230 38.108953 38.635862 38.635862
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.594298 -2.727756 -2.852219 -2.968294 -3.049655 -3.049655
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.4087    1.3137    1.2252     1.1426    1.0656    0.9938
#> (Intercept) 38.635862 38.635862 38.635862 38.8374200 39.241511 39.618368
#> am           0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> cyl         -3.049655 -3.049655 -3.049655 -2.9018833 -2.602618 -2.323531
#> disp         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000 -0.3372238 -1.018971 -1.654754
#>                0.9268    0.8643    0.8061    0.7518    0.7011    0.6538
#> (Intercept) 39.969833 40.252717 40.252451 40.252306 40.252226 40.252182
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.063497 -1.853363 -1.853241 -1.853174 -1.853137 -1.853117
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.247229 -2.725681 -2.725833 -2.725917 -2.725963 -2.725988
#>                0.6098      0.5687     0.5304     0.4946     0.4613     0.4302
#> (Intercept) 40.252157 40.25450212 40.2649799 40.2749246 40.2842067 40.2928637
#> am           0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.000000 -0.01091676 -0.0574394 -0.1007655 -0.1411677 -0.1788468
#> cyl         -1.853106 -1.84844285 -1.8287927 -1.8105785 -1.7935974 -1.7777610
#> disp         0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> vs           0.000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.726002 -2.72548840 -2.7230206 -2.7206151 -2.7183670 -2.7162702
#>                 0.4012     0.3742     0.3489     0.3254     0.3035     0.2830
#> (Intercept) 40.3009372 40.3084665 40.3154884 40.3220371 40.3281444 40.3336890
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb        -0.2139865 -0.2467578 -0.2773205 -0.3058234 -0.3324052 -0.3571341
#> cyl         -1.7629920 -1.7492183 -1.7363729 -1.7243933 -1.7132210 -1.7027647
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.7143148 -2.7124911 -2.7107903 -2.7092041 -2.7077249 -2.7064274
#>                 0.2640     0.2462     0.2296     0.2141     0.1997     0.1862
#> (Intercept) 40.3389992 40.3439623 37.7958811 37.7961553 37.7965044 37.7968507
#> am           0.0000000  0.0000000  1.6596408  1.6597493  1.6597022  1.6595855
#> carb        -0.3802628 -0.4018281 -0.6585303 -0.6585136 -0.6584803 -0.6584427
#> cyl         -1.6930421 -1.6839821 -1.4706309 -1.4708369 -1.4709837 -1.4710856
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.7051424 -2.7039375 -2.2906314 -2.2903488 -2.2901996 -2.2901313
#>                 0.1737     0.1620     0.1510     0.1409     0.1314     0.1225
#> (Intercept) 37.7971569 37.7974096 37.7976083 37.7977588 37.7978693 28.6181866
#> am           1.6594482  1.6593164  1.6592020  1.6591087  1.6590360  2.1882320
#> carb        -0.6584073 -0.6583768 -0.6583522 -0.6583331 -0.6583188 -0.5109634
#> cyl         -1.4711545 -1.4711999 -1.4712288 -1.4712466 -1.4712570 -1.1138595
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.4234982
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.2901096 -2.2901127 -2.2901275 -2.2901462 -2.2901647 -2.6435443
#>                0.1143      0.1066      0.0994      0.0927      0.0864
#> (Intercept)  4.028879 -4.31696955 -4.31452755 -4.31221826 -4.31004833
#> am           1.972490  1.82398575  1.82369603  1.82344769  1.82323564
#> carb        -1.177377 -1.42576039 -1.42595455 -1.42611695 -1.42625191
#> cyl          0.000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         1.191860  1.51389946  1.51363880  1.51340109  1.51318503
#> gear         2.181472  2.47563115  2.47620660  2.47670021  2.47712197
#> hp           0.000000  0.01456208  0.01455867  0.01455539  0.01455226
#> qsec         0.759427  1.07081691  1.07060657  1.07041770  1.07024858
#> vs           0.000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -2.208074 -2.43894810 -2.43857103 -2.43824014 -2.43795058
#>                 0.0806      0.0752      0.0701     0.0654      0.0610
#> (Intercept) -4.3080209 -4.30613613 -4.30439222 -4.3027855 -4.30131092
#> am           1.8230554  1.82290280  1.82277433  1.8226668  1.82257729
#> carb        -1.4263632 -1.42645430 -1.42652805 -1.4265871 -1.42663357
#> cyl          0.0000000  0.00000000  0.00000000  0.0000000  0.00000000
#> disp         0.0000000  0.00000000  0.00000000  0.0000000  0.00000000
#> drat         1.5129893  1.51281245  1.51265319  1.5125102  1.51238206
#> gear         2.4774808  2.47778477  2.47804094  2.4782556  2.47843447
#> hp           0.0145493  0.01454652  0.01454392  0.0145415  0.01453926
#> qsec         1.0700976  1.06996306  1.06984359  1.0697377  1.06964422
#> vs           0.0000000  0.00000000  0.00000000  0.0000000  0.00000000
#> wt          -2.4376979 -2.43747807 -2.43728740 -2.4371226 -2.43698051
#>                 0.0569     0.0530      0.0495      0.0461      0.0430
#> (Intercept) -4.2999627 -4.2987342 -4.29761855 -4.29660850 -9.88431580
#> am           1.8225034  1.8224428  1.82239371  1.82235432  1.60897903
#> carb        -1.4266696 -1.4266967 -1.42671655 -1.42673028 -1.40952900
#> cyl          0.0000000  0.0000000  0.00000000  0.00000000  0.00000000
#> disp         0.0000000  0.0000000  0.00000000  0.00000000  0.00000000
#> drat         1.5122676  1.5121657  1.51207516  1.51199493  1.64688430
#> gear         2.4785824  2.4787037  2.47880237  2.47888164  2.58713168
#> hp           0.0145372  0.0145353  0.01453357  0.01453199  0.01835736
#> qsec         1.0695618  1.0694894  1.06942591  1.06937043  1.37632666
#> vs           0.0000000  0.0000000  0.00000000  0.00000000 -0.72257989
#> wt          -2.4368586 -2.4367542 -2.43666540 -2.43659006 -2.72261817
#>                  0.0401      0.0374      0.0349      0.0325      0.0303
#> (Intercept) -9.89171035 -9.89872479 -9.90537843 -9.91168964 -9.91767588
#> am           1.60867431  1.60838503  1.60811044  1.60784980  1.60760243
#> carb        -1.40960045 -1.40966841 -1.40973305 -1.40979451 -1.40985295
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         1.64706727  1.64724074  1.64740522  1.64756116  1.64770901
#> gear         2.58739906  2.58765314  2.58789454  2.58812386  2.58834170
#> hp           0.01836459  0.01837145  0.01837795  0.01838413  0.01838998
#> qsec         1.37667848  1.37701212  1.37732853  1.37762857  1.37791311
#> vs          -0.72322047 -0.72382804 -0.72440429 -0.72495081 -0.72546914
#> wt          -2.72290182 -2.72317068 -2.72342550 -2.72366704 -2.72389598
#>                  0.0283        0.0264        0.0246        0.0230       0.0214
#> (Intercept) -9.92335372 -13.577070199 -13.581328474 -13.585484046 -13.58953928
#> am           1.60736766   1.731094018   1.731123398   1.731151875   1.73117948
#> carb        -1.40990850  -1.124931806  -1.124766132  -1.124604654  -1.12444727
#> cyl          0.00000000   0.000000000   0.000000000   0.000000000   0.00000000
#> disp         0.00000000   0.009105246   0.009109552   0.009113749   0.00911784
#> drat         1.64784918   1.950109014   1.950306197   1.950498530   1.95068613
#> gear         2.58854860   2.516815719   2.516743135   2.516672502   2.51660377
#> hp           0.01839554   0.012376974   0.012376541   0.012376123   0.01237572
#> qsec         1.37818292   1.567385004   1.567629775   1.567868609   1.56810164
#> vs          -0.72596071  -0.599574650  -0.599765822  -0.599952577  -0.60013501
#> wt          -2.72411298  -3.539696840  -3.540262253  -3.540813629  -3.54135131
#>                    0.0200        0.0186        0.0174        0.0162
#> (Intercept) -13.593496504 -13.597357980 -13.601125931 -13.604802533
#> am            1.731206239   1.731232184   1.731257338   1.731281730
#> carb         -1.124293868  -1.124144355  -1.123998631  -1.123856599
#> cyl           0.000000000   0.000000000   0.000000000   0.000000000
#> disp          0.009121828   0.009125716   0.009129506   0.009133201
#> drat          1.950869104   1.951047571   1.951221636   1.951391406
#> gear          2.516536873   2.516471772   2.516408415   2.516346752
#> hp            0.012375330   0.012374954   0.012374591   0.012374241
#> qsec          1.568329011   1.568550848   1.568767283   1.568978443
#> vs           -0.600313210  -0.600487268  -0.600657272  -0.600823310
#> wt           -3.541875631  -3.542386918  -3.542885489  -3.543371656
#>                    0.0151        0.0141        0.0131        0.0123
#> (Intercept) -13.608389915 -13.611890159 -10.941684786 -10.941161757
#> am            1.731305382   1.731328320   1.592513815   1.592057230
#> carb         -1.123718166  -1.123583241  -1.028730344  -1.028664155
#> cyl           0.000000000   0.000000000  -0.216836005  -0.217156823
#> disp          0.009136802   0.009140313   0.009737391   0.009734404
#> drat          1.951556983   1.951718470   1.814300185   1.814158972
#> gear          2.516286737   2.516228325   2.297293371   2.297009530
#> hp            0.012373903   0.012373576   0.013062494   0.013069857
#> qsec          1.569184453   1.569385433   1.568062200   1.568251265
#> vs           -0.600985466  -0.601143823  -0.780400328  -0.781217577
#> wt           -3.543845721  -3.544307983  -3.662395508  -3.662507454
#>                    0.0114        0.0107       0.0099        0.0093
#> (Intercept) -10.940715802 -10.940340473 -10.94002976 -10.939778072
#> am            1.591616585   1.591191420   1.59078128   1.590385735
#> carb         -1.028606130  -1.028555831  -1.02851284  -1.028476755
#> cyl          -0.217461606  -0.217751021  -0.21802571  -0.218286284
#> disp          0.009731441   0.009728503   0.00972559   0.009722705
#> drat          1.814023962   1.813894854   1.81377136   1.813653220
#> gear          2.296744873   2.296498440   2.29626932   2.296056639
#> hp            0.013077024   0.013083999   0.01309078   0.013097384
#> qsec          1.568432875   1.568607190   1.56877437   1.568934589
#> vs           -0.782004587  -0.782762234  -0.78349138  -0.784192858
#> wt           -3.662606046  -3.662691866  -3.66276548  -3.662827434
#>                    0.0086        0.0081        0.0075        0.0070
#> (Intercept) -10.939580209 -10.939431333 -10.939326951 -10.939262889
#> am            1.590004344   1.589636688   1.589282355   1.588940946
#> carb         -1.028447195  -1.028423795  -1.028406206  -1.028394097
#> cyl          -0.218533337  -0.218767436  -0.218989127  -0.219198934
#> disp          0.009719847   0.009717018   0.009714218   0.009711448
#> drat          1.813540167   1.813431963   1.813328377   1.813229191
#> gear          2.295859569   2.295677318   2.295509130   2.295354284
#> hp            0.013103800   0.013110037   0.013116097   0.013121983
#> qsec          1.569088005   1.569234789   1.569375109   1.569509134
#> vs           -0.784867503  -0.785516120  -0.786139500  -0.786738418
#> wt           -3.662878259  -3.662918471  -3.662948569  -3.662969034
#>                    0.0065        0.0061        0.0057        0.0053
#> (Intercept) -10.939235273 -10.939240511 -10.939275273 -10.939336474
#> am            1.588612065   1.588295332   1.587990372   1.587696820
#> carb         -1.028387148  -1.028385059  -1.028387539  -1.028394314
#> cyl          -0.219397361  -0.219584892  -0.219761991  -0.219929107
#> disp          0.009708708   0.009705999   0.009703321   0.009700676
#> drat          1.813134200   1.813043206   1.812956024   1.812872479
#> gear          2.295212094   2.295081904   2.294963091   2.294855057
#> hp            0.013127700   0.013133251   0.013138638   0.013143866
#> qsec          1.569637034   1.569758978   1.569875132   1.569985664
#> vs           -0.787313630  -0.787865877  -0.788395882  -0.788904351
#> wt           -3.662980334  -3.662982922  -3.662977235  -3.662963695
#>                    0.0049
#> (Intercept) -10.939421262
#> am            1.587414322
#> carb         -1.028405119
#> cyl          -0.220086667
#> disp          0.009698062
#> drat          1.812792402
#> gear          2.294757234
#> hp            0.013148938
#> qsec          1.570090740
#> vs           -0.789391974
#> wt           -3.662942711
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1  25  29  29  28  16   1   1   1   1   1   5   5   5   5   5
#>  [37]   5   5   5   5   5   4   4   4  33   1   1   1   1   1   1   1   1  45
#>  [55] 125  53   1   1   1   1   1   1   1   1   1   1   1   1  85   1   1   1
#>  [73]   1   1   1 165   1   1   1   1   1   1   1   1   1  84   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.946072020 4.612723320 4.301841207 4.011911507 3.741522099 3.489356032
#>   [7] 3.254185114 3.034863928 2.830324257 2.639569875 2.461671700 2.295763266
#>  [13] 2.141036506 1.996737811 1.862164365 1.736660719 1.619615599 1.510458929
#>  [19] 1.408659052 1.313720147 1.225179806 1.142606787 1.065598914 0.993781114
#>  [25] 0.926803593 0.864340132 0.806086500 0.751758968 0.701092930 0.653841613
#>  [31] 0.609774876 0.568678091 0.530351091 0.494607202 0.461272332 0.430184121
#>  [37] 0.401191151 0.374152210 0.348935603 0.325418511 0.303486393 0.283032426
#>  [43] 0.263956987 0.246167170 0.229576327 0.214103651 0.199673781 0.186216437
#>  [49] 0.173666072 0.161961560 0.151045893 0.140865905 0.131372015 0.122517981
#>  [55] 0.114260679 0.106559891 0.099378111 0.092680359 0.086434013 0.080608650
#>  [61] 0.075175897 0.070109293 0.065384161 0.060977488 0.056867809 0.053035109
#>  [67] 0.049460720 0.046127233 0.043018412 0.040119115 0.037415221 0.034893560
#>  [73] 0.032541851 0.030348639 0.028303243 0.026395699 0.024616717 0.022957633
#>  [79] 0.021410365 0.019967378 0.018621644 0.017366607 0.016196156 0.015104589
#>  [85] 0.014086591 0.013137201 0.012251798 0.011426068 0.010655989 0.009937811
#>  [91] 0.009268036 0.008643401 0.008060865 0.007517590 0.007010929 0.006538416
#>  [97] 0.006097749 0.005686781 0.005303511 0.004946072
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 21
#> 
#> $loss
#>   [1] 672.36571 573.74390 491.23208 422.51192 365.58184 318.71483 280.42170
#>   [8] 249.41927 224.60270 205.02153 189.85884 178.41323 170.08313 164.35323
#>  [15] 160.78275 158.99517 158.62952 158.62952 158.62952 158.62952 158.62952
#>  [22] 143.55311 119.14598 103.70624  95.67899  93.67525  93.67525  93.67525
#>  [29]  93.67525  93.67525  93.67525  93.23131  91.47202  90.02674  88.84683
#>  [36]  87.89241  87.12928  86.52800  86.06328  85.71340  85.45976  85.28675
#>  [43]  85.17986  85.12801  80.94081  80.94081  80.94081  80.94081  80.94081
#>  [50]  80.94081  80.94081  80.94081  80.94081  79.19410  73.75972  71.47848
#>  [57]  71.47847  71.47846  71.47845  71.47844  71.47844  71.47843  71.47843
#>  [64]  71.47843  71.47843  71.47843  71.47843  71.47843  71.14227  71.14225
#>  [71]  71.14224  71.14222  71.14221  71.14220  71.14219  70.50134  70.50132
#>  [78]  70.50131  70.50129  70.50128  70.50126  70.50125  70.50124  70.50123
#>  [85]  70.50122  70.39708  70.39706  70.39704  70.39702  70.39701  70.39699
#>  [92]  70.39698  70.39696  70.39695  70.39694  70.39693  70.39692  70.39691
#>  [99]  70.39690  70.39689
#> 
#> $linear.predictors
#>      4.9461   4.6127   4.3018   4.0119   3.7415   3.4894   3.2542   3.0349
#> 1  18.88571 19.03253 19.16944 19.29713 19.41622 19.52727 19.63085 19.72744
#> 2  18.88571 19.03253 19.16944 19.29713 19.41622 19.52727 19.63085 19.72744
#> 3  18.88571 18.41592 17.97778 17.56918 17.18811 16.83272 16.50129 16.19219
#> 4  18.88571 19.03253 19.16944 19.29713 19.41622 19.52727 19.63085 19.72744
#> 5  18.88571 18.41592 17.97778 17.56918 17.18811 16.83272 16.50129 16.19219
#> 6  18.88571 19.64914 20.36111 21.02509 21.64432 22.22183 22.76040 23.26268
#> 7  18.88571 19.03253 19.16944 19.29713 19.41622 19.52727 19.63085 19.72744
#> 8  18.88571 19.03253 19.16944 19.29713 19.41622 19.52727 19.63085 19.72744
#> 9  18.88571 18.41592 17.97778 17.56918 17.18811 16.83272 16.50129 16.19219
#> 10 18.88571 18.41592 17.97778 17.56918 17.18811 16.83272 16.50129 16.19219
#> 11 18.88571 18.41592 17.97778 17.56918 17.18811 16.83272 16.50129 16.19219
#> 12 18.88571 18.41592 17.97778 17.56918 17.18811 16.83272 16.50129 16.19219
#> 13 18.88571 19.64914 20.36111 21.02509 21.64432 22.22183 22.76040 23.26268
#> 14 18.88571 18.41592 17.97778 17.56918 17.18811 16.83272 16.50129 16.19219
#> 15 18.88571 18.41592 17.97778 17.56918 17.18811 16.83272 16.50129 16.19219
#> 16 18.88571 18.41592 17.97778 17.56918 17.18811 16.83272 16.50129 16.19219
#> 17 18.88571 19.64914 20.36111 21.02509 21.64432 22.22183 22.76040 23.26268
#> 18 18.88571 19.64914 20.36111 21.02509 21.64432 22.22183 22.76040 23.26268
#> 19 18.88571 19.03253 19.16944 19.29713 19.41622 19.52727 19.63085 19.72744
#> 20 18.88571 18.41592 17.97778 17.56918 17.18811 16.83272 16.50129 16.19219
#> 21 18.88571 19.64914 20.36111 21.02509 21.64432 22.22183 22.76040 23.26268
#>      2.8303   2.6396   2.4617   2.2958   2.1410   1.9967   1.8622   1.7367
#> 1  19.81752 19.90153 19.97988 20.05295 20.12109 20.18465 20.24391 20.29919
#> 2  19.81752 19.90153 19.97988 20.05295 20.12109 20.18465 20.24391 20.29919
#> 3  15.90393 15.63510 15.38438 15.15056 14.93250 14.72913 14.53948 14.36260
#> 4  19.81752 19.90153 19.97988 20.05295 20.12109 20.18465 20.24391 20.29919
#> 5  15.90393 15.63510 15.38438 15.15056 14.93250 14.72913 14.53948 14.36260
#> 6  23.73111 24.16797 24.57539 24.95534 25.30969 25.64016 25.94835 26.23578
#> 7  19.81752 19.90153 19.97988 20.05295 20.12109 20.18465 20.24391 20.29919
#> 8  19.81752 19.90153 19.97988 20.05295 20.12109 20.18465 20.24391 20.29919
#> 9  15.90393 15.63510 15.38438 15.15056 14.93250 14.72913 14.53948 14.36260
#> 10 15.90393 15.63510 15.38438 15.15056 14.93250 14.72913 14.53948 14.36260
#> 11 15.90393 15.63510 15.38438 15.15056 14.93250 14.72913 14.53948 14.36260
#> 12 15.90393 15.63510 15.38438 15.15056 14.93250 14.72913 14.53948 14.36260
#> 13 23.73111 24.16797 24.57539 24.95534 25.30969 25.64016 25.94835 26.23578
#> 14 15.90393 15.63510 15.38438 15.15056 14.93250 14.72913 14.53948 14.36260
#> 15 15.90393 15.63510 15.38438 15.15056 14.93250 14.72913 14.53948 14.36260
#> 16 15.90393 15.63510 15.38438 15.15056 14.93250 14.72913 14.53948 14.36260
#> 17 23.73111 24.16797 24.57539 24.95534 25.30969 25.64016 25.94835 26.23578
#> 18 23.73111 24.16797 24.57539 24.95534 25.30969 25.64016 25.94835 26.23578
#> 19 19.81752 19.90153 19.97988 20.05295 20.12109 20.18465 20.24391 20.29919
#> 20 15.90393 15.63510 15.38438 15.15056 14.93250 14.72913 14.53948 14.36260
#> 21 23.73111 24.16797 24.57539 24.95534 25.30969 25.64016 25.94835 26.23578
#>      1.6196   1.5105   1.4087   1.3137   1.2252   1.1426   1.0656   0.9938
#> 1  20.33793 20.33793 20.33793 20.33793 20.33793 20.45660 20.69626 20.91977
#> 2  20.33793 20.33793 20.33793 20.33793 20.33793 20.34195 20.34981 20.35715
#> 3  14.23862 14.23862 14.23862 14.23862 14.23862 14.46230 14.91531 15.33777
#> 4  20.33793 20.33793 20.33793 20.33793 20.33793 20.25933 20.10016 19.95173
#> 5  14.23862 14.23862 14.23862 14.23862 14.23862 14.41846 14.78284 15.12265
#> 6  26.43724 26.43724 26.43724 26.43724 26.43724 26.15414 25.58052 25.04558
#> 7  20.33793 20.33793 20.33793 20.33793 20.33793 20.26607 20.12054 19.98483
#> 8  20.33793 20.33793 20.33793 20.33793 20.33793 20.26607 20.12054 19.98483
#> 9  14.23862 14.23862 14.23862 14.23862 14.23862 14.36451 14.61981 14.85789
#> 10 14.23862 14.23862 14.23862 14.23862 14.23862 13.85193 13.07097 12.34266
#> 11 14.23862 14.23862 14.23862 14.23862 14.23862 13.79325 12.89367 12.05474
#> 12 14.23862 14.23862 14.23862 14.23862 14.23862 13.81989 12.97417 12.18546
#> 13 26.43724 26.43724 26.43724 26.43724 26.43724 26.48799 26.58930 26.68379
#> 14 14.23862 14.23862 14.23862 14.23862 14.23862 14.43533 14.83379 15.20539
#> 15 14.23862 14.23862 14.23862 14.23862 14.23862 14.46399 14.92040 15.34604
#> 16 14.23862 14.23862 14.23862 14.23862 14.23862 14.32741 14.50772 14.67587
#> 17 26.43724 26.43724 26.43724 26.43724 26.43724 26.50823 26.65044 26.78307
#> 18 26.43724 26.43724 26.43724 26.43724 26.43724 26.71967 27.28934 27.82060
#> 19 20.33793 20.33793 20.33793 20.33793 20.33793 20.49201 20.80325 21.09351
#> 20 14.23862 14.23862 14.23862 14.23862 14.23862 14.41846 14.78284 15.12265
#> 21 26.43724 26.43724 26.43724 26.43724 26.43724 26.29240 25.99830 25.72403
#>      0.9268   0.8643   0.8061   0.7518   0.7011   0.6538   0.6098   0.5687
#> 1  21.12806 21.29621 21.29623 21.29625 21.29626 21.29626 21.29626 21.28440
#> 2  20.36401 20.36948 20.36945 20.36944 20.36943 20.36943 20.36942 20.39048
#> 3  15.73139 16.04947 16.04966 16.04976 16.04981 16.04984 16.04986 16.06945
#> 4  19.81344 19.70168 19.70162 19.70159 19.70157 19.70156 19.70155 19.72274
#> 5  15.43925 15.69513 15.69530 15.69539 15.69544 15.69547 15.69548 15.69330
#> 6  24.54718 24.14434 24.14408 24.14393 24.14385 24.14381 24.14379 24.14459
#> 7  19.85838 19.75620 19.75614 19.75611 19.75609 19.75608 19.75607 19.74450
#> 8  19.85838 19.75620 19.75614 19.75611 19.75609 19.75608 19.75607 19.74450
#> 9  15.07969 15.25902 15.25916 15.25924 15.25928 15.25931 15.25932 15.26814
#> 10 11.66390 11.11599 11.11590 11.11585 11.11582 11.11581 11.11580 11.11448
#> 11 11.27288 10.64172 10.64160 10.64154 10.64150 10.64148 10.64147 10.64024
#> 12 11.45041 10.85705 10.85694 10.85689 10.85685 10.85684 10.85683 10.85556
#> 13 26.77194 26.84277 26.84265 26.84259 26.84256 26.84254 26.84253 26.85374
#> 14 15.55161 15.83142 15.83159 15.83168 15.83174 15.83177 15.83178 15.85141
#> 15 15.74262 16.06310 16.06329 16.06339 16.06344 16.06347 16.06349 16.08307
#> 16 14.83249 14.95920 14.95932 14.95939 14.95943 14.95945 14.95946 14.95742
#> 17 26.90677 27.00631 27.00620 27.00615 27.00612 27.00610 27.00609 27.00635
#> 18 28.31579 28.71531 28.71530 28.71530 28.71529 28.71529 28.71529 28.71523
#> 19 21.36402 21.58240 21.58245 21.58247 21.58248 21.58249 21.58249 21.54874
#> 20 15.43925 15.69513 15.69530 15.69539 15.69544 15.69547 15.69548 15.64963
#> 21 25.46855 25.26187 25.26167 25.26156 25.26150 25.26147 25.26145 25.26204
#>      0.5304   0.4946   0.4613   0.4302   0.4012   0.3742   0.3489   0.3254
#> 1  21.23378 21.18662 21.14265 21.10163 21.06338 21.02771 20.99445 20.96342
#> 2  20.48027 20.56391 20.64190 20.71464 20.78248 20.84574 20.90474 20.95976
#> 3  16.15257 16.22985 16.30191 16.36911 16.43179 16.49024 16.54475 16.59558
#> 4  19.81313 19.89736 19.97590 20.04916 20.11747 20.18118 20.24060 20.29601
#> 5  15.68370 15.67464 15.66619 15.65830 15.65095 15.64410 15.63770 15.63174
#> 6  24.14849 24.15232 24.15589 24.15922 24.16233 24.16523 24.16793 24.17046
#> 7  19.69528 19.64948 19.60677 19.56694 19.52980 19.49516 19.46285 19.43272
#> 8  19.69528 19.64948 19.60677 19.56694 19.52980 19.49516 19.46285 19.43272
#> 9  15.30545 15.34011 15.37242 15.40255 15.43065 15.45686 15.48130 15.50409
#> 10 11.10902 11.10401 11.09933 11.09497 11.09090 11.08711 11.08357 11.08028
#> 11 10.63522 10.63062 10.62633 10.62234 10.61861 10.61514 10.61190 10.60887
#> 12 10.85034 10.84555 10.84108 10.83692 10.83304 10.82942 10.82605 10.82290
#> 13 26.90172 26.94649 26.98824 27.02718 27.06349 27.09736 27.12894 27.15839
#> 14 15.93473 16.01220 16.08444 16.15181 16.21464 16.27324 16.32788 16.37885
#> 15 16.16618 16.24345 16.31550 16.38269 16.44536 16.50380 16.55830 16.60913
#> 16 14.94848 14.94007 14.93223 14.92491 14.91809 14.91172 14.90579 14.90025
#> 17 27.00767 27.00896 27.01018 27.01131 27.01236 27.01335 27.01426 27.01512
#> 18 28.71500 28.71479 28.71459 28.71441 28.71424 28.71408 28.71393 28.71379
#> 19 21.40482 21.27076 21.14574 21.02915 20.92041 20.81901 20.72444 20.63624
#> 20 15.45394 15.27158 15.10152 14.94292 14.79501 14.65706 14.52842 14.40845
#> 21 25.26493 25.26777 25.27042 25.27289 25.27520 25.27735 25.27936 25.28123
#>      0.3035   0.2830   0.2640   0.2462   0.2296   0.2141   0.1997   0.1862
#> 1  20.93449 20.90759 20.88241 20.85894 21.41205 21.41208 21.41206 21.41202
#> 2  21.01108 21.05880 21.10345 21.14508 20.94919 20.94915 20.94913 20.94912
#> 3  16.64299 16.68719 16.72845 16.76690 16.83400 16.83363 16.83339 16.83323
#> 4  20.34768 20.39573 20.44069 20.48262 20.38798 20.38801 20.38803 20.38804
#> 5  15.62618 15.62109 15.61625 15.61174 15.21916 15.21886 15.21870 15.21863
#> 6  24.17281 24.17486 24.17690 24.17882 23.28918 23.28957 23.28987 23.29010
#> 7  19.40462 19.37845 19.35401 19.33121 18.45820 18.45828 18.45839 18.45851
#> 8  19.40462 19.37845 19.35401 19.33121 18.45820 18.45828 18.45839 18.45851
#> 9  15.52535 15.54519 15.56369 15.58093 15.51119 15.51092 15.51075 15.51065
#> 10 11.07720 11.07429 11.07161 11.06912 11.37090 11.37107 11.37117 11.37121
#> 11 10.60606 10.60337 10.60092 10.59864 10.97233 10.97255 10.97267 10.97272
#> 12 10.81997 10.81718 10.81462 10.81225 11.15329 11.15349 11.15360 11.15364
#> 13 27.18586 27.21136 27.23525 27.25754 27.87508 27.87528 27.87535 27.87536
#> 14 16.42637 16.47068 16.51204 16.55059 16.65075 16.65041 16.65017 16.65002
#> 15 16.65653 16.70073 16.74197 16.78042 16.84545 16.84509 16.84484 16.84468
#> 16 14.89509 14.89035 14.88586 14.88167 14.60069 14.60047 14.60035 14.60029
#> 17 27.01592 27.01661 27.01730 27.01795 27.35399 27.35418 27.35428 27.35433
#> 18 28.71366 28.71354 28.71342 28.71332 28.79021 28.79023 28.79024 28.79024
#> 19 20.55399 20.47749 20.40593 20.33919 20.33551 20.33554 20.33557 20.33560
#> 20 14.29656 14.19255 14.09520 14.00442 14.24468 14.24456 14.24448 14.24444
#> 21 25.28297 25.28449 25.28601 25.28743 25.88798 25.88836 25.88856 25.88864
#>      0.1737   0.1620   0.1510   0.1409   0.1314   0.1225   0.1143   0.1066
#> 1  21.41198 21.41195 21.41191 21.41189 21.41186 21.68716 21.24324 20.42586
#> 2  20.94912 20.94912 20.94912 20.94913 20.94913 21.15788 20.73115 20.92427
#> 3  16.83313 16.83307 16.83303 16.83302 16.83301 16.79953 17.30257 17.41087
#> 4  20.38804 20.38804 20.38804 20.38804 20.38804 20.84054 20.40113 20.60470
#> 5  15.21860 15.21860 15.21861 15.21863 15.21865 14.93422 13.83616 14.08890
#> 6  23.29027 23.29040 23.29048 23.29054 23.29058 23.17788 22.94276 22.85927
#> 7  18.45862 18.45872 18.45879 18.45884 18.45889 18.54740 19.01909 18.81410
#> 8  18.45862 18.45872 18.45879 18.45884 18.45889 18.80150 19.47474 19.45659
#> 9  15.51059 15.51056 15.51055 15.51054 15.51054 15.76757 15.82997 15.85059
#> 10 11.37122 11.37121 11.37120 11.37119 11.37117 11.39935 11.41805 11.27664
#> 11 10.97274 10.97273 10.97272 10.97270 10.97268 10.87161 10.99576 10.93253
#> 12 11.15366 11.15365 11.15364 11.15362 11.15361 10.91105 11.14056 11.26351
#> 13 27.87534 27.87530 27.87526 27.87523 27.87520 28.26973 28.34095 28.60471
#> 14 16.64992 16.64986 16.64982 16.64981 16.64980 16.52452 16.54719 16.10066
#> 15 16.84458 16.84452 16.84449 16.84447 16.84446 16.93133 17.52625 17.35884
#> 16 14.60027 14.60027 14.60028 14.60029 14.60031 14.03835 13.53319 13.75716
#> 17 27.35434 27.35433 27.35432 27.35430 27.35429 26.74429 27.79106 27.72867
#> 18 28.79024 28.79023 28.79023 28.79022 28.79022 28.48649 28.54079 28.79324
#> 19 20.33563 20.33565 20.33567 20.33568 20.33569 20.29909 19.81375 19.20107
#> 20 14.24442 14.24441 14.24441 14.24441 14.24441 14.55346 14.91371 15.64347
#> 21 25.88867 25.88866 25.88864 25.88861 25.88858 25.85707 25.25794 25.50433
#>      0.0994   0.0927   0.0864   0.0806   0.0752   0.0701   0.0654   0.0610
#> 1  20.42565 20.42549 20.42536 20.42527 20.42520 20.42515 20.42512 20.42510
#> 2  20.92418 20.92411 20.92405 20.92399 20.92393 20.92389 20.92384 20.92381
#> 3  17.41095 17.41102 17.41107 17.41112 17.41116 17.41119 17.41121 17.41123
#> 4  20.60465 20.60460 20.60456 20.60452 20.60449 20.60446 20.60443 20.60441
#> 5  14.08864 14.08840 14.08819 14.08801 14.08784 14.08770 14.08758 14.08747
#> 6  22.85944 22.85960 22.85974 22.85986 22.85996 22.86005 22.86013 22.86020
#> 7  18.81408 18.81406 18.81404 18.81404 18.81403 18.81403 18.81403 18.81404
#> 8  19.45644 19.45631 19.45619 19.45609 19.45601 19.45594 19.45587 19.45582
#> 9  15.85047 15.85036 15.85027 15.85019 15.85012 15.85006 15.85001 15.84997
#> 10 11.27677 11.27688 11.27698 11.27707 11.27715 11.27722 11.27729 11.27734
#> 11 10.93270 10.93286 10.93299 10.93310 10.93320 10.93329 10.93336 10.93342
#> 12 11.26363 11.26372 11.26380 11.26387 11.26392 11.26396 11.26400 11.26402
#> 13 28.60441 28.60415 28.60392 28.60371 28.60353 28.60336 28.60322 28.60310
#> 14 16.10099 16.10128 16.10155 16.10179 16.10200 16.10219 16.10236 16.10252
#> 15 17.35895 17.35904 17.35913 17.35920 17.35927 17.35933 17.35939 17.35943
#> 16 13.75695 13.75677 13.75659 13.75644 13.75630 13.75618 13.75607 13.75597
#> 17 27.72914 27.72955 27.72990 27.73020 27.73045 27.73067 27.73086 27.73102
#> 18 28.79353 28.79378 28.79399 28.79416 28.79431 28.79444 28.79454 28.79462
#> 19 19.20117 19.20128 19.20139 19.20150 19.20160 19.20170 19.20179 19.20188
#> 20 15.64316 15.64287 15.64261 15.64237 15.64216 15.64197 15.64179 15.64164
#> 21 25.50409 25.50387 25.50368 25.50351 25.50335 25.50321 25.50309 25.50298
#>      0.0569   0.0530   0.0495   0.0461   0.0430   0.0401   0.0374   0.0349
#> 1  20.42510 20.42511 20.42512 20.42514 20.47478 20.47455 20.47433 20.47412
#> 2  20.92377 20.92374 20.92372 20.92369 20.83926 20.83924 20.83922 20.83920
#> 3  17.41125 17.41126 17.41127 17.41127 17.51752 17.51764 17.51775 17.51786
#> 4  20.60439 20.60437 20.60435 20.60434 20.62696 20.62705 20.62714 20.62722
#> 5  14.08737 14.08729 14.08722 14.08716 14.10428 14.10432 14.10436 14.10440
#> 6  22.86026 22.86031 22.86035 22.86039 22.97911 22.97926 22.97940 22.97953
#> 7  18.81404 18.81405 18.81405 18.81406 18.63823 18.63805 18.63787 18.63771
#> 8  19.45578 19.45574 19.45571 19.45568 19.46402 19.46405 19.46408 19.46411
#> 9  15.84994 15.84991 15.84989 15.84987 16.07674 16.07693 16.07711 16.07728
#> 10 11.27739 11.27743 11.27747 11.27750 11.28020 11.28018 11.28015 11.28013
#> 11 10.93348 10.93352 10.93356 10.93359 10.88511 10.88506 10.88502 10.88498
#> 12 11.26404 11.26406 11.26407 11.26407 11.20381 11.20380 11.20379 11.20377
#> 13 28.60299 28.60289 28.60281 28.60274 28.67927 28.67938 28.67948 28.67958
#> 14 16.10265 16.10277 16.10287 16.10296 15.99204 15.99183 15.99163 15.99144
#> 15 17.35948 17.35952 17.35955 17.35958 17.45757 17.45761 17.45764 17.45768
#> 16 13.75589 13.75581 13.75574 13.75569 13.63373 13.63364 13.63355 13.63347
#> 17 27.73115 27.73126 27.73135 27.73143 27.96573 27.96596 27.96618 27.96639
#> 18 28.79469 28.79475 28.79479 28.79483 28.54242 28.54230 28.54218 28.54207
#> 19 19.20196 19.20204 19.20211 19.20218 19.16882 19.16862 19.16844 19.16826
#> 20 15.64150 15.64138 15.64127 15.64117 15.73840 15.73866 15.73890 15.73914
#> 21 25.50289 25.50280 25.50273 25.50267 25.33200 25.33188 25.33177 25.33166
#>      0.0325   0.0303   0.0283   0.0264   0.0246   0.0230   0.0214   0.0200
#> 1  20.47392 20.47372 20.47354 20.64556 20.64565 20.64574 20.64583 20.64592
#> 2  20.83919 20.83917 20.83916 21.05567 21.05578 21.05588 21.05599 21.05609
#> 3  17.51796 17.51805 17.51814 17.81055 17.81072 17.81089 17.81105 17.81121
#> 4  20.62730 20.62737 20.62745 20.42461 20.42457 20.42453 20.42449 20.42445
#> 5  14.10443 14.10446 14.10449 14.23441 14.23453 14.23465 14.23476 14.23488
#> 6  22.97966 22.97977 22.97989 22.99584 22.99585 22.99587 22.99588 22.99590
#> 7  18.63755 18.63740 18.63726 18.59031 18.59021 18.59011 18.59002 18.58992
#> 8  19.46413 19.46415 19.46417 19.53074 19.53079 19.53083 19.53088 19.53092
#> 9  16.07744 16.07759 16.07774 15.70741 15.70734 15.70728 15.70721 15.70715
#> 10 11.28011 11.28009 11.28008 11.62060 11.62074 11.62088 11.62101 11.62114
#> 11 10.88495 10.88491 10.88488 10.90492 10.90489 10.90485 10.90481 10.90477
#> 12 11.20376 11.20376 11.20375 11.00968 11.00954 11.00941 11.00927 11.00914
#> 13 28.67967 28.67976 28.67984 28.71634 28.71643 28.71652 28.71661 28.71670
#> 14 15.99127 15.99110 15.99094 15.83988 15.83972 15.83957 15.83942 15.83927
#> 15 17.45771 17.45774 17.45777 17.44780 17.44781 17.44782 17.44784 17.44785
#> 16 13.63339 13.63332 13.63325 13.52772 13.52764 13.52757 13.52749 13.52742
#> 17 27.96659 27.96678 27.96696 27.94927 27.94924 27.94921 27.94918 27.94915
#> 18 28.54197 28.54187 28.54177 28.63833 28.63826 28.63820 28.63813 28.63807
#> 19 19.16809 19.16793 19.16778 19.02364 19.02354 19.02343 19.02333 19.02323
#> 20 15.73936 15.73957 15.73977 15.77610 15.77624 15.77637 15.77651 15.77663
#> 21 25.33156 25.33146 25.33137 25.15063 25.15051 25.15040 25.15029 25.15019
#>      0.0186   0.0174   0.0162   0.0151   0.0141   0.0131   0.0123   0.0114
#> 1  20.64600 20.64608 20.64616 20.64624 20.64632 20.65372 20.65367 20.65362
#> 2  21.05619 21.05628 21.05638 21.05647 21.05656 21.08574 21.08566 21.08557
#> 3  17.81137 17.81152 17.81167 17.81181 17.81195 17.75427 17.75411 17.75395
#> 4  20.42441 20.42438 20.42434 20.42431 20.42427 20.44432 20.44446 20.44460
#> 5  14.23499 14.23510 14.23520 14.23530 14.23540 14.39361 14.39385 14.39408
#> 6  22.99591 22.99593 22.99594 22.99595 22.99597 23.15360 23.15394 23.15427
#> 7  18.58983 18.58974 18.58965 18.58957 18.58948 18.49878 18.49862 18.49846
#> 8  19.53096 19.53100 19.53104 19.53108 19.53111 19.43962 19.43957 19.43952
#> 9  15.70710 15.70704 15.70698 15.70693 15.70688 15.67320 15.67348 15.67375
#> 10 11.62127 11.62139 11.62151 11.62163 11.62174 11.65653 11.65639 11.65626
#> 11 10.90474 10.90470 10.90467 10.90463 10.90460 10.90916 10.90907 10.90899
#> 12 11.00902 11.00889 11.00877 11.00866 11.00854 10.98974 10.98973 10.98972
#> 13 28.71678 28.71686 28.71694 28.71702 28.71709 28.66723 28.66724 28.66724
#> 14 15.83913 15.83899 15.83886 15.83872 15.83860 15.78296 15.78275 15.78256
#> 15 17.44786 17.44787 17.44788 17.44789 17.44790 17.33978 17.33966 17.33954
#> 16 13.52735 13.52729 13.52722 13.52716 13.52709 13.57656 13.57665 13.57673
#> 17 27.94912 27.94909 27.94907 27.94904 27.94902 27.95905 27.95915 27.95925
#> 18 28.63801 28.63795 28.63789 28.63784 28.63778 28.63314 28.63286 28.63259
#> 19 19.02313 19.02304 19.02294 19.02285 19.02276 19.08965 19.08974 19.08981
#> 20 15.77676 15.77688 15.77700 15.77711 15.77723 15.72123 15.72127 15.72131
#> 21 25.15008 25.14998 25.14988 25.14979 25.14969 25.17810 25.17813 25.17816
#>      0.0107   0.0099   0.0093   0.0086   0.0081   0.0075   0.0070   0.0065
#> 1  20.65356 20.65350 20.65344 20.65338 20.65331 20.65325 20.65318 20.65311
#> 2  21.08549 21.08541 21.08532 21.08524 21.08516 21.08508 21.08500 21.08493
#> 3  17.75380 17.75365 17.75351 17.75338 17.75324 17.75312 17.75299 17.75287
#> 4  20.44474 20.44487 20.44500 20.44512 20.44524 20.44536 20.44547 20.44558
#> 5  14.39429 14.39450 14.39468 14.39486 14.39503 14.39518 14.39533 14.39547
#> 6  23.15458 23.15488 23.15517 23.15544 23.15570 23.15595 23.15619 23.15642
#> 7  18.49831 18.49817 18.49803 18.49790 18.49778 18.49766 18.49755 18.49744
#> 8  19.43947 19.43943 19.43939 19.43936 19.43932 19.43929 19.43926 19.43923
#> 9  15.67402 15.67428 15.67454 15.67479 15.67503 15.67527 15.67550 15.67573
#> 10 11.65613 11.65601 11.65588 11.65575 11.65563 11.65551 11.65539 11.65527
#> 11 10.90892 10.90884 10.90877 10.90871 10.90864 10.90858 10.90852 10.90846
#> 12 10.98971 10.98971 10.98971 10.98971 10.98972 10.98973 10.98974 10.98975
#> 13 28.66725 28.66726 28.66726 28.66727 28.66728 28.66729 28.66729 28.66730
#> 14 15.78238 15.78220 15.78203 15.78187 15.78172 15.78158 15.78144 15.78131
#> 15 17.33943 17.33932 17.33922 17.33913 17.33904 17.33896 17.33888 17.33881
#> 16 13.57681 13.57688 13.57695 13.57702 13.57708 13.57715 13.57720 13.57726
#> 17 27.95934 27.95944 27.95954 27.95963 27.95972 27.95982 27.95991 27.96000
#> 18 28.63233 28.63208 28.63184 28.63161 28.63139 28.63118 28.63098 28.63078
#> 19 19.08989 19.08996 19.09003 19.09009 19.09015 19.09020 19.09026 19.09030
#> 20 15.72135 15.72139 15.72143 15.72147 15.72150 15.72154 15.72158 15.72162
#> 21 25.17819 25.17822 25.17824 25.17827 25.17829 25.17832 25.17834 25.17837
#>      0.0061   0.0057   0.0053   0.0049
#> 1  20.65304 20.65297 20.65290 20.65283
#> 2  21.08485 21.08477 21.08470 21.08462
#> 3  17.75276 17.75264 17.75254 17.75243
#> 4  20.44569 20.44580 20.44590 20.44600
#> 5  14.39559 14.39571 14.39582 14.39592
#> 6  23.15664 23.15684 23.15704 23.15723
#> 7  18.49734 18.49725 18.49716 18.49707
#> 8  19.43920 19.43917 19.43915 19.43913
#> 9  15.67594 15.67616 15.67637 15.67657
#> 10 11.65516 11.65504 11.65493 11.65482
#> 11 10.90841 10.90836 10.90831 10.90826
#> 12 10.98976 10.98978 10.98979 10.98981
#> 13 28.66731 28.66732 28.66732 28.66733
#> 14 15.78118 15.78107 15.78096 15.78085
#> 15 17.33874 17.33867 17.33861 17.33856
#> 16 13.57731 13.57737 13.57742 13.57746
#> 17 27.96009 27.96017 27.96026 27.96034
#> 18 28.63059 28.63041 28.63024 28.63007
#> 19 19.09035 19.09040 19.09044 19.09048
#> 20 15.72165 15.72169 15.72173 15.72176
#> 21 25.17839 25.17841 25.17843 25.17846
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.4 18.7 18.1 14.3 24.4 19.2 17.8 17.3 10.4 10.4 14.7 32.4 15.5 15.2
#> [16] 13.3 26.0 30.4 19.7 15.0 21.4
#> 
#> $X
#>               am        carb        cyl        disp       drat       gear
#>  [1,]  1.4142136  0.50469494 -0.2936101 -0.77664262  0.8919622  0.4291975
#>  [2,] -0.7071068 -1.26173735 -0.2936101  0.03642611 -0.8435846 -0.8583951
#>  [3,] -0.7071068 -0.67292658  0.9395524  0.88268132 -0.6954281 -0.8583951
#>  [4,] -0.7071068 -1.26173735 -0.2936101 -0.23736234 -1.5208711 -0.8583951
#>  [5,] -0.7071068  0.50469494  0.9395524  0.88268132 -0.5684369 -0.8583951
#>  [6,] -0.7071068 -0.67292658 -1.5267726 -0.88698767  0.4474929  0.4291975
#>  [7,] -0.7071068  0.50469494 -0.2936101 -0.71358831  0.9342926  0.4291975
#>  [8,] -0.7071068  0.50469494 -0.2936101 -0.71358831  0.9342926  0.4291975
#>  [9,] -0.7071068 -0.08411582  0.9395524  0.18410594 -0.8647498 -0.8583951
#> [10,] -0.7071068  0.50469494  0.9395524  1.81190274 -1.1610626 -0.8583951
#> [11,] -0.7071068  0.50469494  0.9395524  1.71234330 -1.0129062 -0.8583951
#> [12,] -0.7071068  0.50469494  0.9395524  1.54641090 -0.5261065 -0.8583951
#> [13,]  1.4142136 -1.26173735 -1.5267726 -1.45115781  1.2729359  0.4291975
#> [14,] -0.7071068 -0.67292658  0.9395524  0.53422330 -1.5208711 -0.8583951
#> [15,] -0.7071068 -0.67292658  0.9395524  0.41807062 -0.6954281 -0.8583951
#> [16,] -0.7071068  0.50469494  0.9395524  0.79971513  0.5321537 -0.8583951
#> [17,]  1.4142136 -0.67292658 -1.5267726 -1.10601843  2.0137180  1.7167902
#> [18,]  1.4142136 -0.67292658 -1.5267726 -1.31509324  0.6168145  1.7167902
#> [19,]  1.4142136  1.68231646 -0.2936101 -0.90109192  0.2993365  1.7167902
#> [20,]  1.4142136  2.85993799  0.9395524  0.39318076  0.1300148  1.7167902
#> [21,]  1.4142136 -0.67292658 -1.5267726 -1.10021079  1.3364315  0.4291975
#>               hp        qsec         vs            wt
#>  [1,] -0.7138470 -0.36655512 -0.7844645 -0.5834784326
#>  [2,] -0.7138470  1.25470958  1.2747549 -0.2295847859
#>  [3,]  0.2537807 -0.36655512 -0.7844645  0.0046095391
#>  [4,] -0.7882799  1.77726597  1.2747549  0.0254268124
#>  [5,]  1.2958414 -1.15708915 -0.7844645  0.1399218158
#>  [6,] -1.4284028  1.62987827  1.2747549 -0.2556063776
#>  [7,] -0.5203214  0.49097331  1.2747549  0.0046095391
#>  [8,] -0.5203214  0.89293977  1.2747549  0.0046095391
#>  [9,]  0.3282136  0.02201245 -0.7844645  0.3064600024
#> [10,]  0.7003782  0.27659120 -0.7844645  1.8885727759
#> [11,]  0.8492440  0.16940015 -0.7844645  2.0696830539
#> [12,]  1.0725427 -0.09857749 -0.7844645  1.9874548242
#> [13,] -1.3688565  1.27480790  1.2747549 -1.2860614076
#> [14,] -0.1183838 -0.46704674 -0.7844645  0.0878786324
#> [15,] -0.1183838 -0.17897078 -0.7844645 -0.0005947792
#> [16,]  1.2958414 -1.44516511 -0.7844645  0.4209550058
#> [17,] -0.9966920 -0.58093723 -0.7844645 -1.3485132276
#> [18,] -0.6691872 -0.44694842  1.2747549 -2.0011347467
#> [19,]  0.2537807 -1.38487014 -0.7844645 -0.6927691176
#> [20,]  2.6356336 -1.98781983 -0.7844645  0.1399218158
#> [21,] -0.7287335  0.69195654  1.2747549 -0.6823604809
#> attr(,"center")
#>  [1]   0.3333333   3.1428571   6.4761905 253.6095238   3.4785714   3.6666667
#>  [7] 157.9523810  17.5671429   0.3809524   3.4355714
#> attr(,"scale")
#>  [1]   0.4714045   1.6983385   1.6218463 120.5310152   0.4724736   0.7766432
#>  [7]  67.1745957   1.4926619   0.4856209   0.9607406
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 29
#> 
#> $lambda.min
#> [1] 0.7010929
#> 
#> $null.dev
#> [1] 32.01741
#> 
#> $Bias
#> [1] 2.190219
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 9.297373