Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 36.642530 34.316891 29.602215 25.215940 21.577631 18.577959 16.122638
#>   [8] 14.130411 12.531302 11.265102 10.280049  9.531690  8.981891  8.597977
#>  [15]  8.299503  7.953314  7.597980  7.076756  6.519223  6.112730  5.795652
#>  [22]  5.553887  5.375332  5.249616  5.167863  5.122488  5.107019  5.115941
#>  [29]  5.144571  5.188935  5.178246  5.094059  5.033298  5.009432  5.009432
#>  [36]  5.009432  5.009432  5.516927  5.517115  5.524078  5.580689  7.707055
#>  [43] 10.396588 10.473429 10.528071 10.747740 11.144592 15.793376 15.816190
#>  [50] 15.569612 15.586004 15.651227 15.643620 16.287475 16.286383 16.226163
#>  [57] 16.225229 17.111854 17.165715 18.899203 19.130259 19.126072 19.126913
#>  [64] 19.059138 19.080109 19.566034 19.727670 19.733870 19.734795 20.131690
#>  [71] 20.145698 20.146228 20.134516 20.135420 20.291777 20.292721 20.238248
#>  [78] 20.039575 20.852522 20.923538 20.875053 20.824795 20.826358 20.848267
#>  [85] 20.849774 20.851228 20.826081 20.827397 20.828669 20.829899 20.836497
#>  [92] 20.837646 20.838757 20.839831 20.840871 20.841853 20.841695 20.842638
#>  [99] 20.843553 20.844438
#> 
#> $cvse
#>   [1] 11.773570 11.454128 10.234743  8.915848  7.848417  6.982290  6.273999
#>   [8]  5.687048  5.191950  4.765906  4.392122  4.058921  3.758776  3.487418
#>  [15]  3.244923  3.036955  2.859016  2.667205  2.465774  2.319915  2.189623
#>  [22]  2.075111  1.977335  1.897554  1.836928  1.796177  1.775335  1.773638
#>  [29]  1.789568  1.821033  1.831674  1.805590  1.782403  1.767658  1.767658
#>  [36]  1.767658  1.767658  1.769307  1.769321  1.770011  1.773574  3.203742
#>  [43]  5.215570  5.215650  5.217373  5.205230  5.212575  9.474282  9.473836
#>  [50]  8.879024  8.876659  8.880710  8.879743  8.943218  8.942006  8.944044
#>  [57]  8.943026  9.967502  9.968550 10.228613 10.216280 10.217127 10.217583
#>  [64] 10.211711 10.210225 10.313900 10.322688 10.321201 10.321549 10.377997
#>  [71] 10.378004 10.378321 10.381371 10.381622 10.372387 10.372597 10.377457
#>  [78] 10.389787 11.476052 11.473437 11.478109 11.484097 11.485520 11.486298
#>  [85] 11.487633 11.488929 11.492003 11.493227 11.494415 11.495569 11.496266
#>  [92] 11.497355 11.498413 11.499441 11.500440 11.501413 11.502475 11.503394
#>  [99] 11.504287 11.505157
#> 
#> $fold
#>  [1]  7 10 10  2  3  1  9  6  4  7  1  8  3  9  8  5  5  4  6  1  2
#> 
#> $lambda
#>   [1] 5.165622873 4.817477180 4.492795342 4.189995972 3.907604267 3.644244818
#>   [7] 3.398634914 3.169578296 2.955959327 2.756737561 2.570942676 2.397669745
#>  [13] 2.236074828 2.085370869 1.944823852 1.813749233 1.691508605 1.577506587
#>  [19] 1.471187922 1.372034780 1.279564228 1.193325882 1.112899711 1.037893995
#>  [25] 0.967943414 0.902707267 0.841867819 0.785128745 0.732213696 0.682864943
#>  [31] 0.636842132 0.593921103 0.553892809 0.516562287 0.481747718 0.449279534
#>  [37] 0.418999597 0.390760427 0.364424482 0.339863491 0.316957830 0.295595933
#>  [43] 0.275673756 0.257094268 0.239766974 0.223607483 0.208537087 0.194482385
#>  [49] 0.181374923 0.169150861 0.157750659 0.147118792 0.137203478 0.127956423
#>  [55] 0.119332588 0.111289971 0.103789400 0.096794341 0.090270727 0.084186782
#>  [61] 0.078512875 0.073221370 0.068286494 0.063684213 0.059392110 0.055389281
#>  [67] 0.051656229 0.048174772 0.044927953 0.041899960 0.039076043 0.036442448
#>  [73] 0.033986349 0.031695783 0.029559593 0.027567376 0.025709427 0.023976697
#>  [79] 0.022360748 0.020853709 0.019448239 0.018137492 0.016915086 0.015775066
#>  [85] 0.014711879 0.013720348 0.012795642 0.011933259 0.011128997 0.010378940
#>  [91] 0.009679434 0.009027073 0.008418678 0.007851287 0.007322137 0.006828649
#>  [97] 0.006368421 0.005939211 0.005538928 0.005165623
#> 
#> $fit
#> $beta
#>               5.1656     4.8175    4.4928    4.1900    3.9076    3.6442
#> (Intercept) 20.28571 22.0657691 23.725854 25.274055 26.717912 28.064459
#> am           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000 -0.5575615 -1.077545 -1.562484 -2.014739 -2.436514
#>                3.3986    3.1696    2.9560    2.7567    2.5709    2.3977
#> (Intercept) 29.320252 30.491409 31.583634 32.602247 33.552209 34.438146
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.829862 -3.196701 -3.538815 -3.857872 -4.155426 -4.432926
#>                2.2361    2.0854    1.9448        1.8137     1.6915     1.5775
#> (Intercept) 35.264375 36.034918 36.753529 37.4101275537 36.0903218 34.2416943
#> am           0.000000  0.000000  0.000000  0.0000000000  0.0000000  0.0000000
#> carb         0.000000  0.000000  0.000000  0.0000000000  0.0000000  0.0000000
#> cyl          0.000000  0.000000  0.000000  0.0000000000  0.0000000  0.0000000
#> disp         0.000000  0.000000  0.000000  0.0000000000  0.0000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.0000000000  0.0000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.0000000000  0.0000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.0000000000  0.0000000  0.0000000
#> qsec         0.000000  0.000000  0.000000  0.0007245037  0.0975159  0.1974919
#> vs           0.000000  0.000000  0.000000  0.0000000000  0.0000000  0.0000000
#> wt          -4.691723 -4.933078 -5.158166 -5.3679117140 -5.4996970 -5.4837783
#>                1.4712     1.3720     1.2796     1.1933     1.1129     1.0379
#> (Intercept) 32.517636 30.9097731 29.4102749 28.0118379 26.7076509 25.4913617
#> am           0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.290731  0.3776861  0.4587807  0.5344098  0.6049418  0.6707202
#> vs           0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.468932 -5.4550867 -5.4421744 -5.4301323 -5.4189017 -5.4084281
#>                 0.9679     0.9027     0.8419     0.7851    0.7322     0.6829
#> (Intercept) 24.3570463 23.2991800 22.3126104 21.3925322 20.534464 19.7342272
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> qsec         0.7320653  0.7892759  0.8426308  0.8923897  0.938795  0.9820728
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> wt          -5.3986604 -5.3895510 -5.3810555 -5.3731326 -5.365744 -5.3588528
#>                0.6368    0.5939    0.5539    0.5166    0.4817    0.4493
#> (Intercept) 18.987923 18.478106 18.478078 18.478078 18.478078 18.478078
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         1.022434  1.050005  1.050007  1.050007  1.050007  1.050007
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -5.352426 -5.348036 -5.348036 -5.348036 -5.348036 -5.348036
#>                0.4190    0.3908    0.3644    0.3399    0.3170    0.2956
#> (Intercept) 18.478078 18.478078 18.478078 18.478078 18.478078 18.478078
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         1.050007  1.050007  1.050007  1.050007  1.050007  1.050007
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -5.348036 -5.348036 -5.348036 -5.348036 -5.348036 -5.348036
#>                0.2757    0.2571      0.2398      0.2236     0.2085     0.1945
#> (Intercept) 18.478078 18.478078 13.12233849 13.36311710 13.5882448 13.7974170
#> am           0.000000  0.000000  1.48677714  1.55781687  1.6233578  1.6853138
#> carb         0.000000  0.000000 -0.01309347 -0.07582321 -0.1340603 -0.1886691
#> cyl          0.000000  0.000000  0.00000000  0.00000000  0.0000000  0.0000000
#> disp         0.000000  0.000000  0.00000000  0.00000000  0.0000000  0.0000000
#> drat         0.000000  0.000000  0.00000000  0.00000000  0.0000000  0.0000000
#> gear         0.000000  0.000000  0.00000000  0.00000000  0.0000000  0.0000000
#> hp           0.000000  0.000000  0.00000000  0.00000000  0.0000000  0.0000000
#> qsec         1.050007  1.050007  1.20849996  1.19279294  1.1781611  1.1645010
#> vs           0.000000  0.000000  0.00000000  0.00000000  0.0000000  0.0000000
#> wt          -5.348036 -5.348036 -4.75125929 -4.69253980 -4.6381897 -4.5870269
#>                 0.1814      0.1692      0.1578      0.1471      0.1372
#> (Intercept) 13.9924600 18.93393541 19.10072591 19.19721247 19.19876859
#> am           1.7431277  1.90213634  1.95130136  1.98054640  1.98080739
#> carb        -0.2396094 -0.28064976 -0.32454395 -0.35018175 -0.35020222
#> cyl          0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.0000000 -0.01619409 -0.01618249 -0.01617901 -0.01618659
#> qsec         1.1517610  0.87275836  0.86190731  0.85557296  0.85546681
#> vs           0.0000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -4.5392932 -3.82208512 -3.78166779 -3.75772116 -3.75730997
#>                  0.1280      0.1193      0.1113      0.1038      0.0968
#> (Intercept) 19.20008029 19.20117835 19.20209245 19.20284998 19.20347541
#> am           1.98101161  1.98117200  1.98129842  1.98139835  1.98147758
#> carb        -0.35021734 -0.35022856 -0.35023694 -0.35024324 -0.35024799
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.01619285 -0.01619799 -0.01620222 -0.01620568 -0.01620851
#> qsec         0.85537862  0.85530565  0.85524547  0.85519600  0.85515542
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -3.75697595 -3.75670476 -3.75648468 -3.75630615 -3.75616137
#>                  0.0903     0.0842      0.0785      0.0732      0.0683
#> (Intercept) 19.20399020 19.2044128 19.20475906 18.21870035 18.22083580
#> am           1.98154055  1.9815907  1.98163074  1.68385251  1.68374044
#> carb        -0.35025160 -0.3502544 -0.35025647 -0.41752583 -0.41752589
#> cyl          0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> drat         0.00000000  0.0000000  0.00000000  0.56387857  0.56398959
#> gear         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> hp          -0.01621082 -0.0162127 -0.01621423 -0.01668587 -0.01669312
#> qsec         0.85512219  0.8550950  0.85507289  0.78055375  0.78041706
#> vs           0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> wt          -3.75604400 -3.7559489 -3.75587176 -3.55654030 -3.55624792
#>                  0.0637      0.0594     0.0554      0.0517      0.0482
#> (Intercept) 18.22252961 18.22387209 18.2249358 18.22577878 18.22644716
#> am           1.68366208  1.68360786  1.6835708  1.68354569  1.68352900
#> carb        -0.41752601 -0.41752630 -0.4175268 -0.41752742 -0.41752819
#> cyl          0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> drat         0.56407474  0.56414061  0.5641920  0.56423259  0.56426486
#> gear         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> hp          -0.01669895 -0.01670363 -0.0167074 -0.01671042 -0.01671285
#> qsec         0.78030853  0.78022227  0.7801536  0.78009893  0.78005525
#> vs           0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> wt          -3.55601011 -3.55581666 -3.5556592 -3.55553112 -3.55542678
#>                  0.0449      0.0419      0.0391      0.0364      0.0340
#> (Intercept) 18.22697761 18.22739912 17.08354136 17.08001126 17.07704457
#> am           1.68351804  1.68351099  1.60305010  1.60288580  1.60276632
#> carb        -0.41752904 -0.41752994 -0.41740644 -0.41733050 -0.41726891
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         0.56429078  0.56431178  0.68620936  0.68614462  0.68607486
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.01671481 -0.01671638 -0.01691855 -0.01691069 -0.01690382
#> qsec         0.78002032  0.77999231  0.83558722  0.83582981  0.83603570
#> vs           0.00000000  0.00000000 -0.33577768 -0.33615429 -0.33645835
#> wt          -3.55534177 -3.55527246 -3.58601321 -3.58652314 -3.58695892
#>                  0.0317      0.0296      0.0276        0.0257       0.0240
#> (Intercept) 17.07453889 17.07241561 17.07061279 17.0718279264 31.855463697
#> am           1.60267987  1.60261758  1.60257278  1.6025405639  0.397222286
#> carb        -0.41721887 -0.41717819 -0.41714508 -0.4171181165 -0.063883342
#> cyl          0.00000000  0.00000000  0.00000000 -0.0001639602 -1.089039658
#> disp         0.00000000  0.00000000  0.00000000  0.0000000000  0.003935745
#> drat         0.68600617  0.68594199  0.68588405  0.6854650178  0.825625437
#> gear         0.00000000  0.00000000  0.00000000  0.0000000000 -1.340709265
#> hp          -0.01689783 -0.01689265 -0.01688818 -0.0168820900  0.000000000
#> qsec         0.83621058  0.83635914  0.83648528  0.8365602080  0.618472103
#> vs          -0.33670567 -0.33690814 -0.33707481 -0.3371599207 -0.698230845
#> wt          -3.58733013 -3.58764547 -3.58791278 -3.5881839383 -4.572790217
#>                   0.0224       0.0209       0.0194       0.0181      0.0169
#> (Intercept) 31.856891570 31.857773676 31.858189852 31.858211370 31.85790161
#> am           0.396359668  0.395601415  0.394938672  0.394362968  0.39386626
#> carb        -0.063953405 -0.064025083 -0.064097655 -0.064170509 -0.06424313
#> cyl         -1.089168654 -1.089262002 -1.089324413 -1.089360144 -1.08937302
#> disp         0.003933213  0.003930815  0.003928549  0.003926409  0.00392439
#> drat         0.826524703  0.827337784  0.828070518  0.828728673  0.82931788
#> gear        -1.340969448 -1.341178953 -1.341343821 -1.341469551 -1.34156114
#> hp           0.000000000  0.000000000  0.000000000  0.000000000  0.00000000
#> qsec         0.618322086  0.618196415  0.618092273  0.618007092  0.61793853
#> vs          -0.698737254 -0.699188507 -0.699588922 -0.699942667 -0.70025373
#> wt          -4.572453049 -4.572140106 -4.571849694 -4.571580192 -4.57133006
#>                   0.0158       0.0147       0.0137       0.0128       0.0119
#> (Intercept) 31.857316714 31.856506187 31.855513512 31.854376697 31.853128803
#> am           0.393440968  0.393079969  0.392776628  0.392524782  0.392318734
#> carb        -0.064315073 -0.064385988 -0.064455574 -0.064523591 -0.064589850
#> cyl         -1.089366469 -1.089343550 -1.089306985 -1.089259182 -1.089202267
#> disp         0.003922486  0.003920692  0.003919003  0.003917412  0.003915914
#> drat         0.829843609  0.830311088  0.830725320  0.831091044  0.831412726
#> gear        -1.341623095 -1.341659500 -1.341674010 -1.341669904 -1.341650106
#> hp           0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> qsec         0.617884481  0.617843026  0.617812450  0.617791214  0.617777944
#> vs          -0.700525917 -0.700762811 -0.700967790 -0.701144013 -0.701294423
#> wt          -4.571097851 -4.570882186 -4.570681783 -4.570495441 -4.570322040
#>                   0.0111       0.0104       0.0097       0.0090       0.0084
#> (Intercept) 31.851798440 31.850410217 31.848985171 31.847541155 31.846093197
#> am           0.392153240  0.392023496  0.391925113  0.391854101  0.391806843
#> carb        -0.064654204 -0.064716542 -0.064776788 -0.064834892 -0.064890828
#> cyl         -1.089138103 -1.089068322 -1.088994340 -1.088917384 -1.088838506
#> disp         0.003914504  0.003913178  0.003911928  0.003910752  0.003909644
#> drat         0.831694555  0.831940440  0.832154015  0.832338640  0.832497416
#> gear        -1.341617214 -1.341573529 -1.341521081 -1.341461652 -1.341396797
#> hp           0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> qsec         0.617771420  0.617770558  0.617774403  0.617782116  0.617792961
#> vs          -0.701421745 -0.701528498 -0.701616996 -0.701689360 -0.701747526
#> wt          -4.570160541 -4.570009980 -4.569869467 -4.569738182 -4.569615368
#>                  0.0079       0.0073       0.0068      0.0064       0.0059
#> (Intercept) 31.84465383 31.843233371 31.841840237 31.84048114 31.839161339
#> am           0.39178007  0.391770859  0.391776564  0.39179484  0.391823600
#> carb        -0.06494459 -0.064996191 -0.065045653 -0.06509301 -0.065138321
#> cyl         -1.08875861 -1.088678444 -1.088598663 -1.08851979 -1.088442272
#> disp         0.00390860  0.003907615  0.003906687  0.00390581  0.003904983
#> drat         0.83263319  0.832748558  0.832845902  0.83292738  0.832994940
#> gear        -1.34132787 -1.341256046 -1.341182326 -1.34110757 -1.341032512
#> hp           0.00000000  0.000000000  0.000000000  0.00000000  0.000000000
#> qsec         0.61780630  0.617821572  0.617838305  0.61785608  0.617874561
#> vs          -0.70179325 -0.701828133 -0.701853609 -0.70187097 -0.701881390
#> wt          -4.56950033 -4.569392443 -4.569291113 -4.56919581 -4.569106062
#>                   0.0055       0.0052
#> (Intercept) 31.837884825 31.836654504
#> am           0.391860990  0.391905380
#> carb        -0.065181625 -0.065222987
#> cyl         -1.088366452 -1.088292614
#> disp         0.003904201  0.003903462
#> drat         0.833050351  0.833095198
#> gear        -1.340957759 -1.340883821
#> hp           0.000000000  0.000000000
#> qsec         0.617893440  0.617912473
#> vs          -0.701885895 -0.701885410
#> wt          -4.569021414 -4.568941468
#> 
#> $iter
#>   [1]   0   3   3   2   2   2   2   2   2   2   2   2   2   2   2   3   3   3
#>  [19]   3   3   3   3   3   3   3   3   3   3   3   3   3   2   1   1   1   1
#>  [37]   1   1   1   1   1   1   1   1 120  16  15  15  15  50  19  14   1   1
#>  [55]   1   1   1   1   1   1   1  36   1   1   1   1   1   1   1   1  27   1
#>  [73]   1   1   1   1   1 272   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.165622873 4.817477180 4.492795342 4.189995972 3.907604267 3.644244818
#>   [7] 3.398634914 3.169578296 2.955959327 2.756737561 2.570942676 2.397669745
#>  [13] 2.236074828 2.085370869 1.944823852 1.813749233 1.691508605 1.577506587
#>  [19] 1.471187922 1.372034780 1.279564228 1.193325882 1.112899711 1.037893995
#>  [25] 0.967943414 0.902707267 0.841867819 0.785128745 0.732213696 0.682864943
#>  [31] 0.636842132 0.593921103 0.553892809 0.516562287 0.481747718 0.449279534
#>  [37] 0.418999597 0.390760427 0.364424482 0.339863491 0.316957830 0.295595933
#>  [43] 0.275673756 0.257094268 0.239766974 0.223607483 0.208537087 0.194482385
#>  [49] 0.181374923 0.169150861 0.157750659 0.147118792 0.137203478 0.127956423
#>  [55] 0.119332588 0.111289971 0.103789400 0.096794341 0.090270727 0.084186782
#>  [61] 0.078512875 0.073221370 0.068286494 0.063684213 0.059392110 0.055389281
#>  [67] 0.051656229 0.048174772 0.044927953 0.041899960 0.039076043 0.036442448
#>  [73] 0.033986349 0.031695783 0.029559593 0.027567376 0.025709427 0.023976697
#>  [79] 0.022360748 0.020853709 0.019448239 0.018137492 0.016915086 0.015775066
#>  [85] 0.014711879 0.013720348 0.012795642 0.011933259 0.011128997 0.010378940
#>  [91] 0.009679434 0.009027073 0.008418678 0.007851287 0.007322137 0.006828649
#>  [97] 0.006368421 0.005939211 0.005538928 0.005165623
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 44
#> 
#> $loss
#>   [1] 694.82571 587.25414 497.25453 422.29813 360.20175 309.08164 267.31347
#>   [8] 233.49763 206.42900 185.07087 168.53220 156.04791 146.96186 140.71199
#>  [15] 136.81749 134.77316 122.29099 111.03723 101.66527  93.90209  87.51200
#>  [22]  82.29175  78.06623  74.68465  72.01731  69.95272  68.39517  67.26256
#>  [29]  66.48458  66.00108  65.76069  65.71328  65.71328  65.71328  65.71328
#>  [36]  65.71328  65.71328  65.71328  65.71328  65.71328  65.71328  65.71328
#>  [43]  65.71328  65.71328  61.77241  61.04597  60.50372  60.11085  59.84524
#>  [50]  56.37392  56.29543  56.28301  56.28300  56.28300  56.28300  56.28299
#>  [57]  56.28299  56.28299  56.28299  56.28299  56.28299  55.76054  55.76053
#>  [64]  55.76053  55.76053  55.76052  55.76052  55.76052  55.76052  55.76052
#>  [71]  55.61506  55.61504  55.61503  55.61503  55.61502  55.61502  55.61471
#>  [78]  54.34807  54.34806  54.34804  54.34803  54.34802  54.34802  54.34801
#>  [85]  54.34801  54.34800  54.34800  54.34800  54.34799  54.34799  54.34799
#>  [92]  54.34799  54.34799  54.34799  54.34798  54.34798  54.34798  54.34798
#>  [99]  54.34798  54.34798
#> 
#> $linear.predictors
#>      5.1656   4.8175   4.4928   4.1900   3.9076   3.6442   3.3986   3.1696
#> 1  20.28571 20.60496 20.90269 21.18035 21.43930 21.68079 21.90601 22.11605
#> 2  20.28571 20.46278 20.62791 20.78191 20.92554 21.05948 21.18440 21.30090
#> 3  20.28571 20.27321 20.26155 20.25067 20.24053 20.23107 20.22224 20.21402
#> 4  20.28571 20.14776 20.01910 19.89911 19.78721 19.68285 19.58553 19.49476
#> 5  20.28571 20.28715 20.28849 20.28973 20.29090 20.29198 20.29299 20.29393
#> 6  20.28571 20.30945 20.33159 20.35223 20.37148 20.38944 20.40619 20.42180
#> 7  20.28571 20.14776 20.01910 19.89911 19.78721 19.68285 19.58553 19.49476
#> 8  20.28571 20.14776 20.01910 19.89911 19.78721 19.68285 19.58553 19.49476
#> 9  20.28571 19.79649 19.34025 18.91475 18.51792 18.14785 17.80271 17.48084
#> 10 20.28571 19.98606 19.70661 19.44599 19.20294 18.97626 18.76487 18.56772
#> 11 20.28571 19.13857 18.06874 17.07102 16.14053 15.27276 14.46347 13.70873
#> 12 20.28571 19.04156 17.88125 16.79914 15.78997 14.84881 13.97108 13.15251
#> 13 20.28571 21.16531 21.98562 22.75064 23.46411 24.12949 24.75002 25.32874
#> 14 20.28571 21.04264 21.74856 22.40690 23.02087 23.59346 24.12745 24.62546
#> 15 20.28571 20.10315 19.93290 19.77411 19.62603 19.48793 19.35914 19.23902
#> 16 20.28571 20.15055 20.02449 19.90692 19.79728 19.69503 19.59967 19.51074
#> 17 20.28571 20.98689 21.64080 22.25065 22.81939 23.34980 23.84447 24.30579
#> 18 20.28571 20.87259 21.41991 21.93034 22.40637 22.85032 23.26435 23.65047
#> 19 20.28571 20.29830 20.31004 20.32098 20.33119 20.34071 20.34959 20.35787
#> 20 20.28571 20.52132 20.74105 20.94598 21.13709 21.31532 21.48153 21.63655
#> 21 20.28571 20.51575 20.73028 20.93035 21.11694 21.29095 21.45323 21.60458
#>      2.9560   2.7567   2.5709   2.3977   2.2361    2.0854    1.9448    1.8137
#> 1  22.31194 22.49462 22.66499 22.82388 22.97206 23.110254 23.239133 23.358124
#> 2  21.40954 21.51086 21.60536 21.69348 21.77567 21.852319 21.923801 21.989712
#> 3  20.20634 20.19919 20.19251 20.18629 20.18049 20.175072 20.170024 20.166376
#> 4  19.41011 19.33117 19.25754 19.18888 19.12485 19.065130 19.009437 18.956842
#> 5  20.29481 20.29563 20.29640 20.29711 20.29778 20.298399 20.298978 20.300979
#> 6  20.43637 20.44995 20.46262 20.47443 20.48545 20.495722 20.505305 20.517797
#> 7  19.41011 19.33117 19.25754 19.18888 19.12485 19.065130 19.009437 18.957770
#> 8  19.41011 19.33117 19.25754 19.18888 19.12485 19.065130 19.009437 18.958204
#> 9  17.18066 16.90071 16.63962 16.39614 16.16906 15.957291 15.759792 15.575333
#> 10 18.38385 18.21238 18.05247 17.90333 17.76425 17.634537 17.513568 17.400568
#> 11 13.00485 12.34842 11.73622 11.16529 10.63283 10.136259  9.673155  9.241618
#> 12 12.38910 11.67715 11.01318 10.39396  9.81647  9.277903  8.775634  8.307485
#> 13 25.86845 26.37178 26.84120 27.27897 27.68724 28.067997 28.423090 28.754368
#> 14 25.08991 25.52305 25.92700 26.30373 26.65506 26.982720 27.288294 27.574427
#> 15 19.12701 19.02254 18.92511 18.83425 18.74951 18.670484 18.596783 18.527301
#> 16 19.42780 19.35046 19.27832 19.21105 19.14831 19.089795 19.035227 18.983885
#> 17 24.73603 25.13726 25.51146 25.86044 26.18589 26.489412 26.772477 27.036912
#> 18 24.01057 24.34640 24.65960 24.95169 25.22409 25.478131 25.715053 25.934896
#> 19 20.36559 20.37279 20.37951 20.38577 20.39161 20.397061 20.402141 20.404353
#> 20 21.78112 21.91594 22.04168 22.15894 22.26830 22.370292 22.465408 22.552242
#> 21 21.74573 21.87736 22.00012 22.11461 22.22139 22.320961 22.413826 22.500809
#>       1.6915    1.5775    1.4712    1.3720    1.2796    1.1933    1.1129
#> 1  23.286227 23.124911 22.974465 22.834159 22.703309 22.581277 22.467471
#> 2  21.938413 21.837143 21.742697 21.654616 21.572472 21.495863 21.424418
#> 3  20.304505 20.450589 20.586829 20.713887 20.832382 20.942890 21.045951
#> 4  18.831085 18.738808 18.652750 18.572492 18.497643 18.427839 18.362739
#> 5  20.496606 20.698279 20.886362 21.061768 21.225353 21.377913 21.520191
#> 6  20.999390 21.490356 21.948239 22.375261 22.773504 23.144907 23.491278
#> 7  18.955905 18.991598 19.024886 19.055930 19.084882 19.111883 19.137064
#> 8  19.014415 19.110093 19.199324 19.282542 19.360151 19.432529 19.500029
#> 9  15.403332 15.359075 15.317801 15.279308 15.243410 15.209931 15.178708
#> 10 17.292732 17.263058 17.235384 17.209575 17.185505 17.163058 17.142123
#> 11  8.970248  9.002762  9.033085  9.061364  9.087737  9.112333  9.135271
#> 12  7.997698  8.016986  8.034974  8.051749  8.067394  8.081984  8.095591
#> 13 29.014306 29.042942 29.069648 29.094554 29.117782 29.139445 29.159647
#> 14 27.938944 28.109049 28.267692 28.415642 28.553621 28.682301 28.802308
#> 15 18.376482 18.270482 18.171626 18.079432 17.993452 17.913267 17.838485
#> 16 18.885888 18.821525 18.761500 18.705520 18.653313 18.604624 18.559217
#> 17 27.291459 27.363179 27.430067 27.492447 27.550623 27.604878 27.655476
#> 18 25.949486 25.804523 25.669328 25.543245 25.425660 25.315999 25.213730
#> 19 20.070263 19.721749 19.396720 19.093597 18.810903 18.547261 18.301389
#> 20 22.367658 22.112752 21.875024 21.653317 21.446553 21.253724 21.073891
#> 21 22.614960 22.670139 22.721600 22.769593 22.814352 22.856093 22.895022
#>       1.0379    0.9679    0.9027    0.8419    0.7851    0.7322    0.6829
#> 1  22.361334 22.262351 22.170039 22.083948 22.003659 21.928782 21.858951
#> 2  21.357788 21.295649 21.237698 21.183652 21.133249 21.086243 21.042405
#> 3  21.142065 21.231702 21.315298 21.393260 21.465967 21.533774 21.597011
#> 4  18.302026 18.245406 18.192601 18.143356 18.097429 18.054598 18.014653
#> 5  21.652879 21.776625 21.892031 21.999659 22.100033 22.193643 22.280943
#> 6  23.814305 24.115561 24.396514 24.658531 24.902889 25.130778 25.343308
#> 7  19.160548 19.182449 19.202874 19.221923 19.239688 19.256255 19.271706
#> 8  19.562980 19.621689 19.676440 19.727501 19.775122 19.819532 19.860950
#> 9  15.149590 15.122435 15.097109 15.073490 15.051463 15.030921 15.011763
#> 10 17.122600 17.104392 17.087412 17.071575 17.056806 17.043033 17.030188
#> 11  9.156663  9.176613  9.195219  9.212571  9.228753  9.243845  9.257919
#> 12  8.108281  8.120116  8.131153  8.141446  8.151046  8.159998  8.168347
#> 13 29.178488 29.196059 29.212446 29.227728 29.241980 29.255272 29.267668
#> 14 28.914228 29.018604 29.115945 29.206726 29.291389 29.370346 29.443981
#> 15 17.768744 17.703703 17.643046 17.586477 17.533720 17.484519 17.438634
#> 16 18.516870 18.477377 18.440546 18.406197 18.374164 18.344289 18.316427
#> 17 27.702665 27.746672 27.787714 27.825990 27.861686 27.894976 27.926023
#> 18 25.118352 25.029403 24.946449 24.869086 24.796937 24.729650 24.666898
#> 19 18.072087 17.858240 17.658805 17.472811 17.299353 17.137585 16.986720
#> 20 20.906179 20.749769 20.603901 20.467864 20.340995 20.222677 20.112334
#> 21 22.931327 22.965185 22.996761 23.026209 23.053672 23.079285 23.103171
#>       0.6368    0.5939    0.5539    0.5166    0.4817    0.4493    0.4190
#> 1  21.793827 21.749339 21.749336 21.749336 21.749336 21.749336 21.749336
#> 2  21.001521 20.973593 20.973591 20.973591 20.973591 20.973591 20.973591
#> 3  21.655986 21.696273 21.696275 21.696275 21.696275 21.696275 21.696275
#> 4  17.977400 17.951952 17.951951 17.951951 17.951951 17.951951 17.951951
#> 5  22.362360 22.417977 22.417980 22.417980 22.417980 22.417980 22.417980
#> 6  25.541515 25.676914 25.676921 25.676921 25.676921 25.676921 25.676921
#> 7  19.286116 19.295959 19.295960 19.295960 19.295960 19.295960 19.295960
#> 8  19.899576 19.925962 19.925964 19.925964 19.925964 19.925964 19.925964
#> 9  14.993897 14.981691 14.981691 14.981691 14.981691 14.981691 14.981691
#> 10 17.018208 17.010025 17.010024 17.010024 17.010024 17.010024 17.010024
#> 11  9.271045  9.280012  9.280012  9.280012  9.280012  9.280012  9.280012
#> 12  8.176134  8.181453  8.181453  8.181453  8.181453  8.181453  8.181453
#> 13 29.279229 29.287126 29.287127 29.287127 29.287127 29.287127 29.287127
#> 14 29.512654 29.559566 29.559568 29.559568 29.559568 29.559568 29.559568
#> 15 17.395841 17.366609 17.366607 17.366607 17.366607 17.366607 17.366607
#> 16 18.290444 18.272694 18.272693 18.272693 18.272693 18.272693 18.272693
#> 17 27.954977 27.974757 27.974758 27.974758 27.974758 27.974758 27.974758
#> 18 24.608376 24.568398 24.568395 24.568395 24.568395 24.568395 24.568395
#> 19 16.846022 16.749909 16.749903 16.749903 16.749903 16.749903 16.749903
#> 20 20.009427 19.939128 19.939125 19.939125 19.939125 19.939125 19.939125
#> 21 23.125447 23.140664 23.140665 23.140665 23.140665 23.140665 23.140665
#>       0.3908    0.3644    0.3399    0.3170    0.2956    0.2757    0.2571
#> 1  21.749336 21.749336 21.749336 21.749336 21.749336 21.749336 21.749336
#> 2  20.973591 20.973591 20.973591 20.973591 20.973591 20.973591 20.973591
#> 3  21.696275 21.696275 21.696275 21.696275 21.696275 21.696275 21.696275
#> 4  17.951951 17.951951 17.951951 17.951951 17.951951 17.951951 17.951951
#> 5  22.417980 22.417980 22.417980 22.417980 22.417980 22.417980 22.417980
#> 6  25.676921 25.676921 25.676921 25.676921 25.676921 25.676921 25.676921
#> 7  19.295960 19.295960 19.295960 19.295960 19.295960 19.295960 19.295960
#> 8  19.925964 19.925964 19.925964 19.925964 19.925964 19.925964 19.925964
#> 9  14.981691 14.981691 14.981691 14.981691 14.981691 14.981691 14.981691
#> 10 17.010024 17.010024 17.010024 17.010024 17.010024 17.010024 17.010024
#> 11  9.280012  9.280012  9.280012  9.280012  9.280012  9.280012  9.280012
#> 12  8.181453  8.181453  8.181453  8.181453  8.181453  8.181453  8.181453
#> 13 29.287127 29.287127 29.287127 29.287127 29.287127 29.287127 29.287127
#> 14 29.559568 29.559568 29.559568 29.559568 29.559568 29.559568 29.559568
#> 15 17.366607 17.366607 17.366607 17.366607 17.366607 17.366607 17.366607
#> 16 18.272693 18.272693 18.272693 18.272693 18.272693 18.272693 18.272693
#> 17 27.974758 27.974758 27.974758 27.974758 27.974758 27.974758 27.974758
#> 18 24.568395 24.568395 24.568395 24.568395 24.568395 24.568395 24.568395
#> 19 16.749903 16.749903 16.749903 16.749903 16.749903 16.749903 16.749903
#> 20 19.939125 19.939125 19.939125 19.939125 19.939125 19.939125 19.939125
#> 21 23.140665 23.140665 23.140665 23.140665 23.140665 23.140665 23.140665
#>       0.2398    0.2236    0.2085    0.1945    0.1814    0.1692    0.1578
#> 1  22.000352 21.956559 21.915837 21.877731 21.842188 22.283862 22.252803
#> 2  21.465541 21.427925 21.392868 21.360160 21.329655 21.797975 21.771146
#> 3  21.327186 21.388673 21.445857 21.499357 21.549257 21.550354 21.593525
#> 4  17.320489 17.370470 17.417054 17.460514 17.501045 17.245044 17.280428
#> 5  22.109634 22.098127 22.087522 22.077484 22.068117 22.631318 22.622950
#> 6  25.804334 25.744929 25.689716 25.638018 25.589795 24.780795 24.739726
#> 7  18.841182 18.745598 18.656980 18.573737 18.496081 18.642968 18.576071
#> 8  19.566282 19.461274 19.363876 19.272438 19.187137 19.166623 19.093215
#> 9  14.773332 14.791608 14.808635 14.824528 14.839350 14.807159 14.820046
#> 10 16.630460 16.625630 16.621252 16.617018 16.613062 16.281219 16.278194
#> 11  9.854683  9.870407  9.884845  9.898578  9.911397 10.117796 10.128478
#> 12  8.834604  8.863059  8.889294  8.914115  8.937278  9.151171  9.170738
#> 13 29.291064 29.281361 29.272350 29.263903 29.256025 29.423497 29.416580
#> 14 29.926611 29.970880 30.011871 30.050438 30.086420 29.857171 29.888217
#> 15 16.759113 16.816147 16.869275 16.918877 16.965138 17.213216 17.253171
#> 16 17.682625 17.727914 17.770130 17.809509 17.846235 17.913379 17.945233
#> 17 28.242985 28.308833 28.369891 28.427234 28.480729 28.586010 28.631960
#> 18 24.597183 24.646894 24.693047 24.736322 24.776691 25.196912 25.231416
#> 19 17.018499 17.037788 17.055636 17.072444 17.088126 15.977219 15.991444
#> 20 20.101316 19.955950 19.820953 19.694418 19.576385 19.258786 19.157172
#> 21 23.852527 23.909976 23.963112 24.013177 24.059889 24.117525 24.157488
#>       0.1471    0.1372    0.1280    0.1193    0.1113    0.1038    0.0968
#> 1  22.234842 22.235074 22.235264 22.235421 22.235550 22.235655 22.235741
#> 2  21.755744 21.756021 21.756248 21.756433 21.756583 21.756706 21.756806
#> 3  21.618605 21.618565 21.618532 21.618506 21.618485 21.618468 21.618454
#> 4  17.300813 17.300610 17.300444 17.300310 17.300201 17.300112 17.300041
#> 5  22.618079 22.618313 22.618508 22.618670 22.618804 22.618915 22.619005
#> 6  24.715642 24.715302 24.715021 24.714791 24.714602 24.714448 24.714322
#> 7  18.536892 18.536905 18.536922 18.536940 18.536957 18.536972 18.536986
#> 8  19.050235 19.050186 19.050149 19.050123 19.050104 19.050090 19.050079
#> 9  14.827490 14.827447 14.827412 14.827383 14.827361 14.827342 14.827327
#> 10 16.276230 16.276025 16.275859 16.275724 16.275615 16.275526 16.275453
#> 11 10.134954 10.135125 10.135261 10.135371 10.135459 10.135530 10.135587
#> 12  9.182429  9.182612  9.182758  9.182875  9.182969  9.183044  9.183104
#> 13 29.412578 29.412659 29.412725 29.412780 29.412825 29.412862 29.412892
#> 14 29.906425 29.906371 29.906323 29.906281 29.906245 29.906215 29.906189
#> 15 17.276335 17.276370 17.276401 17.276427 17.276451 17.276470 17.276487
#> 16 17.963638 17.963592 17.963556 17.963529 17.963507 17.963491 17.963478
#> 17 28.658901 28.658987 28.659054 28.659107 28.659149 28.659182 28.659209
#> 18 25.251651 25.251844 25.252003 25.252132 25.252237 25.252323 25.252393
#> 19 15.999605 15.999103 15.998687 15.998343 15.998060 15.997828 15.997637
#> 20 19.097835 19.097696 19.097585 19.097495 19.097423 19.097365 19.097318
#> 21 24.181076 24.181194 24.181286 24.181358 24.181413 24.181457 24.181492
#>       0.0903    0.0842    0.0785    0.0732    0.0683    0.0637    0.0594
#> 1  22.235811 22.235868 22.235914 22.125909 22.126084 22.126226 22.126342
#> 2  21.756888 21.756954 21.757008 21.656101 21.656274 21.656417 21.656533
#> 3  21.618443 21.618434 21.618426 21.442162 21.442125 21.442094 21.442069
#> 4  17.299982 17.299935 17.299897 17.290365 17.290260 17.290173 17.290100
#> 5  22.619080 22.619141 22.619191 22.695548 22.695842 22.696077 22.696263
#> 6  24.714219 24.714135 24.714067 24.680474 24.680146 24.679884 24.679673
#> 7  18.536998 18.537009 18.537018 18.756274 18.756457 18.756599 18.756709
#> 8  19.050071 19.050066 19.050061 19.224606 19.224707 19.224784 19.224843
#> 9  14.827315 14.827306 14.827298 14.800289 14.800272 14.800257 14.800244
#> 10 16.275395 16.275347 16.275309 16.165624 16.165480 16.165362 16.165266
#> 11 10.135632 10.135669 10.135699 10.142677 10.142729 10.142774 10.142813
#> 12  9.183153  9.183192  9.183224  9.271563  9.271623  9.271674  9.271719
#> 13 29.412917 29.412937 29.412954 29.691800 29.691934 29.692040 29.692124
#> 14 29.906167 29.906149 29.906134 29.786781 29.786618 29.786490 29.786390
#> 15 17.276501 17.276513 17.276522 17.085993 17.086070 17.086130 17.086177
#> 16 17.963467 17.963459 17.963453 17.943849 17.943886 17.943913 17.943932
#> 17 28.659230 28.659246 28.659260 28.554944 28.554924 28.554911 28.554903
#> 18 25.252450 25.252496 25.252533 25.471320 25.471519 25.471677 25.471804
#> 19 15.997481 15.997354 15.997250 16.250743 16.250266 16.249881 16.249573
#> 20 19.097280 19.097250 19.097225 18.765577 18.765424 18.765304 18.765208
#> 21 24.181519 24.181541 24.181558 24.197400 24.197359 24.197332 24.197313
#>       0.0554    0.0517    0.0482    0.0449    0.0419    0.0391   0.0364
#> 1  22.126436 22.126512 22.126574 22.126624 22.126664 22.190553 22.19043
#> 2  21.656629 21.656707 21.656771 21.656823 21.656865 21.744048 21.74393
#> 3  21.442049 21.442032 21.442018 21.442007 21.441997 21.297625 21.29754
#> 4  17.290040 17.289990 17.289948 17.289914 17.289886 17.335351 17.33552
#> 5  22.696412 22.696530 22.696625 22.696701 22.696761 22.668475 22.66819
#> 6  24.679504 24.679368 24.679259 24.679171 24.679101 24.834635 24.83532
#> 7  18.756795 18.756862 18.756915 18.756956 18.756989 18.642458 18.64225
#> 8  19.224887 19.224922 19.224948 19.224968 19.224984 19.143810 19.14375
#> 9  14.800234 14.800224 14.800217 14.800210 14.800205 14.836789 14.83685
#> 10 16.165188 16.165125 16.165073 16.165031 16.164996 16.223151 16.22343
#> 11 10.142846 10.142875 10.142899 10.142919 10.142936 10.153495 10.15337
#> 12  9.271756  9.271788  9.271816  9.271839  9.271858  9.274684  9.27451
#> 13 29.692191 29.692244 29.692286 29.692320 29.692348 29.702912 29.70275
#> 14 29.786312 29.786250 29.786202 29.786163 29.786133 29.777356 29.77749
#> 15 17.086214 17.086242 17.086264 17.086281 17.086294 17.078474 17.07839
#> 16 17.943946 17.943955 17.943962 17.943967 17.943970 18.010209 18.01025
#> 17 28.554898 28.554895 28.554894 28.554893 28.554893 28.470180 28.47004
#> 18 25.471905 25.471987 25.472053 25.472106 25.472149 25.632336 25.63218
#> 19 16.249325 16.249126 16.248967 16.248839 16.248736 16.194625 16.19494
#> 20 18.765132 18.765070 18.765021 18.764981 18.764948 18.723830 18.72408
#> 21 24.197301 24.197293 24.197289 24.197286 24.197286 24.065005 24.06477
#>       0.0340    0.0317   0.0296    0.0276    0.0257    0.0240    0.0224
#> 1  22.190323 22.190227 22.19014 22.190069 22.189935 22.149077 22.149063
#> 2  21.743828 21.743736 21.74365 21.743583 21.743422 21.329359 21.329348
#> 3  21.297469 21.297419 21.29738 21.297350 21.297488 21.416910 21.416490
#> 4  17.335664 17.335788 17.33589 17.335985 17.336031 17.806837 17.806840
#> 5  22.667943 22.667721 22.66753 22.667357 22.667350 22.716643 22.716890
#> 6  24.835896 24.836378 24.83678 24.837124 24.837450 24.859797 24.859817
#> 7  18.642069 18.641907 18.64176 18.641638 18.641437 18.488348 18.488690
#> 8  19.143691 19.143633 19.14358 19.143529 19.143373 18.859432 18.859683
#> 9  14.836899 14.836942 14.83698 14.837008 14.837002 14.699676 14.699905
#> 10 16.223673 16.223876 16.22405 16.224195 16.224297 16.378119 16.378203
#> 11 10.153274 10.153189 10.15312 10.153059 10.153015 10.255219 10.255067
#> 12  9.274364  9.274242  9.27414  9.274055  9.273983  9.371163  9.371187
#> 13 29.702608 29.702480 29.70237 29.702272 29.701968 30.145009 30.145379
#> 14 29.777612 29.777721 29.77782 29.777902 29.777992 29.452072 29.451752
#> 15 17.078328 17.078274 17.07823 17.078189 17.078213 16.860948 16.860756
#> 16 18.010284 18.010310 18.01033 18.010345 18.010261 17.782472 17.782572
#> 17 28.469927 28.469839 28.46977 28.469713 28.469766 28.291825 28.291544
#> 18 25.632042 25.631916 25.63180 25.631708 25.631658 25.738918 25.739422
#> 19 16.195222 16.195479 16.19571 16.195910 16.195955 15.918975 15.918727
#> 20 18.724298 18.724487 18.72465 18.724787 18.725107 19.110738 19.110304
#> 21 24.064585 24.064436 24.06432 24.064221 24.064298 24.368463 24.368361
#>       0.0209    0.0194    0.0181    0.0169    0.0158    0.0147   0.0137
#> 1  22.149041 22.149012 22.148977 22.148938 22.148896 22.148852 22.14881
#> 2  21.329335 21.329322 21.329308 21.329295 21.329282 21.329269 21.32926
#> 3  21.416110 21.415765 21.415454 21.415174 21.414922 21.414696 21.41449
#> 4  17.806839 17.806835 17.806828 17.806818 17.806805 17.806790 17.80677
#> 5  22.717100 22.717276 22.717423 22.717544 22.717641 22.717719 22.71778
#> 6  24.859851 24.859895 24.859948 24.860008 24.860072 24.860139 24.86021
#> 7  18.488998 18.489275 18.489524 18.489746 18.489944 18.490120 18.49028
#> 8  18.859916 18.860131 18.860328 18.860509 18.860675 18.860826 18.86096
#> 9  14.700119 14.700318 14.700503 14.700674 14.700834 14.700981 14.70112
#> 10 16.378286 16.378365 16.378441 16.378514 16.378584 16.378650 16.37871
#> 11 10.254921 10.254782 10.254651 10.254527 10.254410 10.254300 10.25420
#> 12  9.371201  9.371208  9.371209  9.371205  9.371196  9.371185  9.37117
#> 13 30.145702 30.145982 30.146223 30.146428 30.146602 30.146747 30.14687
#> 14 29.451476 29.451238 29.451037 29.450866 29.450724 29.450607 29.45051
#> 15 16.860583 16.860427 16.860287 16.860162 16.860051 16.859953 16.85987
#> 16 17.782669 17.782761 17.782849 17.782931 17.783009 17.783082 17.78315
#> 17 28.291291 28.291067 28.290868 28.290693 28.290539 28.290405 28.29029
#> 18 25.739866 25.740256 25.740597 25.740893 25.741150 25.741371 25.74156
#> 19 15.918529 15.918376 15.918262 15.918180 15.918128 15.918099 15.91809
#> 20 19.109905 19.109538 19.109203 19.108896 19.108616 19.108361 19.10813
#> 21 24.368263 24.368170 24.368081 24.367998 24.367921 24.367849 24.36778
#>       0.0128    0.0119    0.0111    0.0104    0.0097    0.0090    0.0084
#> 1  22.148759 22.148711 22.148664 22.148617 22.148571 22.148525 22.148481
#> 2  21.329245 21.329235 21.329225 21.329216 21.329208 21.329200 21.329193
#> 3  21.414315 21.414156 21.414014 21.413889 21.413778 21.413681 21.413595
#> 4  17.806753 17.806733 17.806710 17.806687 17.806662 17.806637 17.806611
#> 5  22.717823 22.717854 22.717874 22.717884 22.717886 22.717880 22.717869
#> 6  24.860276 24.860344 24.860412 24.860477 24.860541 24.860602 24.860661
#> 7  18.490414 18.490535 18.490641 18.490733 18.490814 18.490884 18.490945
#> 8  18.861088 18.861201 18.861304 18.861396 18.861479 18.861553 18.861620
#> 9  14.701245 14.701363 14.701471 14.701572 14.701665 14.701752 14.701832
#> 10 16.378772 16.378828 16.378880 16.378930 16.378976 16.379019 16.379060
#> 11 10.254099 10.254008 10.253924 10.253844 10.253771 10.253702 10.253637
#> 12  9.371154  9.371136  9.371117  9.371097  9.371077  9.371057  9.371037
#> 13 30.146964 30.147041 30.147100 30.147143 30.147173 30.147191 30.147199
#> 14 29.450436 29.450377 29.450333 29.450303 29.450283 29.450273 29.450271
#> 15 16.859788 16.859721 16.859662 16.859610 16.859566 16.859527 16.859494
#> 16 17.783212 17.783271 17.783325 17.783375 17.783421 17.783463 17.783501
#> 17 28.290190 28.290105 28.290033 28.289974 28.289924 28.289885 28.289853
#> 18 25.741719 25.741854 25.741966 25.742058 25.742133 25.742193 25.742240
#> 19 15.918102 15.918126 15.918163 15.918209 15.918262 15.918322 15.918386
#> 20 19.107921 19.107732 19.107562 19.107408 19.107270 19.107146 19.107035
#> 21 24.367724 24.367670 24.367622 24.367579 24.367541 24.367508 24.367479
#>       0.0079    0.0073    0.0068   0.0064    0.0059    0.0055   0.0052
#> 1  22.148439 22.148397 22.148358 22.14832 22.148283 22.148248 22.14821
#> 2  21.329188 21.329182 21.329178 21.32917 21.329171 21.329168 21.32917
#> 3  21.413520 21.413455 21.413398 21.41335 21.413306 21.413269 21.41324
#> 4  17.806585 17.806558 17.806531 17.80650 17.806477 17.806450 17.80642
#> 5  22.717853 22.717834 22.717811 22.71779 22.717759 22.717731 22.71770
#> 6  24.860717 24.860769 24.860819 24.86087 24.860909 24.860950 24.86099
#> 7  18.490997 18.491041 18.491079 18.49111 18.491138 18.491160 18.49118
#> 8  18.861680 18.861734 18.861782 18.86182 18.861863 18.861897 18.86193
#> 9  14.701906 14.701976 14.702040 14.70210 14.702155 14.702207 14.70226
#> 10 16.379098 16.379133 16.379166 16.37920 16.379226 16.379253 16.37928
#> 11 10.253578 10.253522 10.253470 10.25342 10.253377 10.253336 10.25330
#> 12  9.371017  9.370997  9.370979  9.37096  9.370943  9.370926  9.37091
#> 13 30.147198 30.147189 30.147175 30.14715 30.147131 30.147103 30.14707
#> 14 29.450276 29.450286 29.450302 29.45032 29.450343 29.450368 29.45039
#> 15 16.859466 16.859441 16.859421 16.85940 16.859390 16.859379 16.85937
#> 16 17.783536 17.783568 17.783598 17.78362 17.783648 17.783670 17.78369
#> 17 28.289829 28.289811 28.289799 28.28979 28.289788 28.289788 28.28979
#> 18 25.742275 25.742300 25.742318 25.74233 25.742332 25.742331 25.74233
#> 19 15.918453 15.918522 15.918593 15.91866 15.918734 15.918803 15.91887
#> 20 19.106936 19.106847 19.106768 19.10670 19.106635 19.106579 19.10653
#> 21 24.367455 24.367434 24.367417 24.36740 24.367392 24.367384 24.36738
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 21.4 18.7 24.4 22.8 19.2 17.8 16.4 17.3 10.4 10.4 30.4 33.9 15.5
#> [16] 15.2 27.3 26.0 15.8 19.7 21.4
#> 
#> $X
#>               am       carb         cyl       disp        drat       gear
#>  [1,]  1.1547005  0.9190183 -0.05643326 -0.5102080  0.40236925  0.3483665
#>  [2,]  1.1547005  0.9190183 -0.05643326 -0.5102080  0.40236925  0.3483665
#>  [3,] -0.8660254 -1.3969078 -0.05643326  0.3169608 -1.03216460 -1.1147728
#>  [4,] -0.8660254 -0.6249324  1.12866530  1.1778916 -0.90970439 -1.1147728
#>  [5,] -0.8660254 -0.6249324 -1.24153183 -0.6224666  0.03498863  0.3483665
#>  [6,] -0.8660254 -0.6249324 -1.24153183 -0.6722656  0.43735788  0.3483665
#>  [7,] -0.8660254  0.9190183 -0.05643326 -0.4460602  0.43735788  0.3483665
#>  [8,] -0.8660254  0.9190183 -0.05643326 -0.4460602  0.43735788  0.3483665
#>  [9,] -0.8660254  0.1470429  1.12866530  0.4672017 -1.04965892 -1.1147728
#> [10,] -0.8660254  0.1470429  1.12866530  0.4672017 -1.04965892 -1.1147728
#> [11,] -0.8660254  0.9190183  1.12866530  2.1232274 -1.29457933 -1.1147728
#> [12,] -0.8660254  0.9190183  1.12866530  2.0219414 -1.17211912 -1.1147728
#> [13,]  1.1547005 -0.6249324 -1.24153183 -1.2217420  2.20428372  0.3483665
#> [14,]  1.1547005 -1.3969078 -1.24153183 -1.2605683  0.96218734  0.3483665
#> [15,] -0.8660254 -0.6249324  1.12866530  0.8233907 -1.59198269 -1.1147728
#> [16,] -0.8660254 -0.6249324  1.12866530  0.7052237 -0.90970439 -1.1147728
#> [17,]  1.1547005 -1.3969078 -1.24153183 -1.1938883  0.71726693  0.3483665
#> [18,]  1.1547005 -0.6249324 -1.24153183 -0.8452958  1.32956796  1.8115058
#> [19,]  1.1547005  0.9190183  1.12866530  1.1019271  0.96218734  1.8115058
#> [20,]  1.1547005  2.4629690 -0.05643326 -0.6368155 -0.08747158  1.8115058
#> [21,]  1.1547005 -0.6249324 -1.24153183 -0.8393874  0.76974987  0.3483665
#>               hp         qsec         vs           wt
#>  [1,] -0.4317458 -0.880138274 -0.8660254 -0.611322161
#>  [2,] -0.4317458 -0.556383939 -0.8660254 -0.339064163
#>  [3,] -0.4317458  0.842697297  1.1547005  0.023946502
#>  [4,]  0.7469202 -0.556383939 -0.8660254  0.264174147
#>  [5,] -1.3021452  1.166451632  1.1547005 -0.002745459
#>  [6,] -0.7037456  2.843036583  1.1547005 -0.045452596
#>  [7,] -0.1960126  0.183625971  1.1547005  0.264174147
#>  [8,] -0.1960126  0.530505616  1.1547005  0.264174147
#>  [9,]  0.8375868 -0.336693497 -0.8660254  0.936811555
#> [10,]  0.8375868 -0.221066948 -0.8660254  0.573800891
#> [11,]  1.2909198 -0.001376507 -0.8660254  2.196672097
#> [12,]  1.4722531 -0.093877745 -0.8660254  2.382448143
#> [13,] -1.4834785  0.310815174  1.1547005 -1.684338979
#> [14,] -1.2477453  1.108638358  1.1547005 -1.449449725
#> [15,]  0.2935871 -0.643103850 -0.8660254  0.349588422
#> [16,]  0.2935871 -0.394506771 -0.8660254  0.258835755
#> [17,] -1.2296119  0.530505616  1.1547005 -1.342681882
#> [18,] -0.7762789 -0.741386416 -0.8660254 -1.123807805
#> [19,]  2.3607859 -2.013278448 -0.8660254 -0.024099027
#> [20,]  0.7469202 -1.435145706 -0.8660254 -0.451170397
#> [21,] -0.4498791  0.357065793  1.1547005 -0.440493613
#> attr(,"center")
#>  [1]   0.4285714   2.8095238   6.0952381 220.4476190   3.6700000   3.7619048
#>  [7] 133.8095238  17.9823810   0.4285714   3.1925714
#> attr(,"scale")
#>  [1]   0.4948717   1.2953781   1.6876233 118.4764217   0.5716143   0.6834619
#>  [7]  55.1470925   1.7297066   0.4948717   0.9366116
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 37
#> 
#> $lambda.min
#> [1] 0.4189996
#> 
#> $null.dev
#> [1] 33.08694
#> 
#> $Bias
#> [1] 1.268405
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 12.24949