Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 50.805824 47.930539 42.409211 37.113751 32.785702 29.280342 26.473063
#>   [8] 24.256663 22.538999 21.240955 20.294670 19.642017 19.233266 19.025941
#>  [15] 18.983817 19.060004 18.735752 18.552606 18.522357 18.522357 18.522357
#>  [22] 18.427923 18.296926 18.179715 18.073679 18.030801 18.008872 17.337405
#>  [29] 16.090294 15.254284 13.435627 12.752875 11.270229 10.798460 10.632356
#>  [36] 10.441928 10.326676 10.028129  9.812491  9.672163  9.615856  9.613952
#>  [43]  9.614028  9.545008  8.383808  8.406699  8.427773  9.187621  9.206518
#>  [50]  9.841594  9.909756 10.965533 11.015404 10.605987 10.621099 10.518615
#>  [57] 10.518986 10.614148 18.053059 17.414118 17.415018 17.415920 17.107229
#>  [64] 17.108131 17.279171 17.205339 17.275057 16.825941 17.422446 17.423179
#>  [71] 17.423924 17.424678 16.900504 16.901494 16.902478 16.903455 16.904426
#>  [78] 16.905391 16.891440 16.892416 16.893384 16.894346 16.895300 16.896248
#>  [85] 16.897188 16.898121 16.899046 16.832143 16.833030 16.833907 16.834775
#>  [92] 16.806096 16.807049 16.807992 16.808924 16.809846 16.810756 16.811656
#>  [99] 16.812545 16.813424
#> 
#> $cvse
#>   [1] 12.932466 12.556922 11.386371 10.047559  8.898449  7.910392  7.061511
#>   [8]  6.335907  5.722799  5.215518  4.810247  4.504441  4.295018  4.176736
#>  [15]  4.141313  4.171730  4.065242  3.999160  3.986714  3.986714  3.986714
#>  [22]  4.002775  3.989375  3.970566  3.946717  3.928567  3.915814  4.431394
#>  [29]  4.095498  4.002296  3.968567  3.860100  3.809195  3.773249  3.754596
#>  [36]  3.747127  3.733251  3.729240  3.739069  3.748533  3.752429  3.752750
#>  [43]  3.752766  3.740093  3.681439  3.684466  3.687315  3.775705  3.778246
#>  [50]  3.865786  3.900736  3.909740  3.920851  3.696714  3.716649  3.724029
#>  [57]  3.724000  3.712852  6.829983  6.827450  6.828333  6.829208  6.826236
#>  [64]  6.827097  6.817446  6.826355  6.821908  6.758187  6.978738  6.979588
#>  [71]  6.980426  6.981251  7.015875  7.016666  7.017446  7.018216  7.018974
#>  [78]  7.019722  7.022120  7.022844  7.023558  7.024263  7.024957  7.025641
#>  [85]  7.026316  7.026981  7.027637  7.031419  7.032062  7.032695  7.033319
#>  [92]  7.035542  7.036147  7.036742  7.037330  7.037908  7.038479  7.039042
#>  [99]  7.039596  7.040143
#> 
#> $fold
#>  [1]  2  5  8  6  8  6  4  7  1  4 10  3  1  9  7  9  2  3  5 10  1
#> 
#> $lambda
#>   [1] 6.151405893 5.736821724 5.350179140 4.989594972 4.653312971 4.339695251
#>   [7] 4.047214315 3.774445616 3.520060614 3.282820310 3.061569208 2.855229690
#>  [13] 2.662796765 2.483333175 2.315964831 2.159876552 2.014308102 1.878550477
#>  [19] 1.751942462 1.633867404 1.523750209 1.421054545 1.325280225 1.235960773
#>  [25] 1.152661154 1.074975650 1.002525889 0.934958999 0.871945892 0.813179657
#>  [31] 0.758374070 0.707262196 0.659595091 0.615140589 0.573682172 0.535017914
#>  [37] 0.498959497 0.465331297 0.433969525 0.404721432 0.377444562 0.352006061
#>  [43] 0.328282031 0.306156921 0.285522969 0.266279677 0.248333318 0.231596483
#>  [49] 0.215987655 0.201430810 0.187855048 0.175194246 0.163386740 0.152375021
#>  [55] 0.142105455 0.132528022 0.123596077 0.115266115 0.107497565 0.100252589
#>  [61] 0.093495900 0.087194589 0.081317966 0.075837407 0.070726220 0.065959509
#>  [67] 0.061514059 0.057368217 0.053501791 0.049895950 0.046533130 0.043396953
#>  [73] 0.040472143 0.037744456 0.035200606 0.032828203 0.030615692 0.028552297
#>  [79] 0.026627968 0.024833332 0.023159648 0.021598766 0.020143081 0.018785505
#>  [85] 0.017519425 0.016338674 0.015237502 0.014210545 0.013252802 0.012359608
#>  [91] 0.011526612 0.010749756 0.010025259 0.009349590 0.008719459 0.008131797
#>  [97] 0.007583741 0.007072622 0.006595951 0.006151406
#> 
#> $fit
#> $beta
#>               6.1514     5.7368    5.3502    4.9896    4.6533    4.3397
#> (Intercept) 19.97143 21.8411725 23.584902 25.211110 26.727717 28.142109
#> am           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000 -0.5592295 -1.080769 -1.567158 -2.020766 -2.443803
#>                4.0472    3.7744    3.5201    3.2828    3.0616    2.8552
#> (Intercept) 29.461177 30.691343 31.838601 32.908537 33.906363 34.836939
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.838328 -3.206264 -3.549402 -3.869414 -4.167858 -4.446188
#>                2.6628    2.4833    2.3160    2.1599    2.0143    1.8786
#> (Intercept) 35.704797 36.514165 37.268984 37.972930 38.466352 38.466352
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -4.705759 -4.947836 -5.173598 -5.384144 -5.531724 -5.531724
#>                1.7519    1.6339    1.5238    1.4211      1.3253     1.2360
#> (Intercept) 38.466352 38.466352 38.466352 38.466352 37.73568106 36.1193304
#> am           0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.03783793  0.1215409
#> vs           0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> wt          -5.531724 -5.531724 -5.531724 -5.531724 -5.51676293 -5.4836673
#>                 1.1527     1.0750     1.0025     0.9350     0.8719     0.8132
#> (Intercept) 34.6123445 33.2069245 31.8962251 30.6738624 27.7396918 18.2178103
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.5242426  2.9924804
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.1995805  0.2723604  0.3402352  0.4035355  0.5193408  0.8422727
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.4528109 -5.4240341 -5.3971968 -5.3721683 -5.1699053 -4.3055080
#>                0.7584     0.7073     0.6596     0.6151    0.5737     0.5350
#> (Intercept) 30.861890 29.1601077 26.9753439 24.9397147 23.039925 21.2681757
#> am           5.975223  6.0209264  6.0734144  6.1229447  6.168474  6.2109243
#> carb        -1.661494 -1.6016451 -1.5229060 -1.4497811 -1.381289 -1.3174092
#> cyl          0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> disp         0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> drat         0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> gear         0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> hp           0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> qsec         0.000000  0.0843231  0.1929836  0.2941863  0.388681  0.4768076
#> vs           0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> wt          -2.456818 -2.4563547 -2.4589267 -2.4609663 -2.463252 -2.4653886
#>                 0.4990     0.4653     0.4340     0.4047     0.3774     0.3520
#> (Intercept) 19.6158369 18.0748602 16.6393402 16.2869964 16.2872091 16.2875299
#> am           6.2505130  6.2874336  6.3224899  6.3306474  6.3300184  6.3295482
#> carb        -1.2578349 -1.2022757 -1.1507621 -1.1378733 -1.1376740 -1.1375422
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.5589948  0.6356429  0.7070026  0.7245457  0.7245759  0.7245903
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.4673817 -2.4692404 -2.4706020 -2.4712489 -2.4715796 -2.4718167
#>                 0.3283     0.3062     0.2855     0.2663     0.2483     0.2316
#> (Intercept) 16.2878289 16.2880709 16.2882547 16.2883900 16.2884879 16.2885579
#> am           6.3292052  6.3289585  6.3287824  6.3286574  6.3285688  6.3285061
#> carb        -1.1374530 -1.1373917 -1.1373491 -1.1373193 -1.1372985 -1.1372838
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.7245974  0.7246013  0.7246034  0.7246047  0.7246055  0.7246061
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.4719856 -2.4721054 -2.4721902 -2.4722501 -2.4722924 -2.4723223
#>                 0.2160     0.2014     0.1879     0.1752    0.1634     0.1524
#> (Intercept) 16.2886078 16.2886431 16.2886681 16.2886858 16.288698 16.2887071
#> am           6.3284619  6.3284306  6.3284086  6.3283930  6.328382  6.3283743
#> carb        -1.1372735 -1.1372662 -1.1372611 -1.1372575 -1.137255 -1.1372532
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> qsec         0.7246064  0.7246067  0.7246068  0.7246069  0.724607  0.7246071
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> wt          -2.4723434 -2.4723583 -2.4723688 -2.4723762 -2.472381 -2.4723851
#>                 0.1421      0.1325      0.1236      0.1153      0.1075
#> (Intercept) 16.2887133 -13.7872317 -13.7917487 -13.7959964 -13.7999910
#> am           6.3283688   6.2222996   6.2223262   6.2223512   6.2223747
#> carb        -1.1372519  -2.1179098  -2.1181211  -2.1183199  -2.1185068
#> cyl          0.0000000   1.9769598   1.9773017   1.9776233   1.9779257
#> disp         0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> drat         0.0000000   1.8863043   1.8865728   1.8868252   1.8870626
#> gear         0.0000000   2.2827465   2.2831531   2.2835354   2.2838950
#> hp           0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> qsec         0.7246071   0.9019034   0.9019012   0.9018991   0.9018971
#> vs           0.0000000   2.3330295   2.3335171   2.3339757   2.3344070
#> wt          -2.4723877  -2.0716980  -2.0715875  -2.0714835  -2.0713857
#>                  0.1003      0.0935      0.0872        0.0813        0.0758
#> (Intercept) -13.8037476 -13.8072804 -13.8106026 -31.404336999 -31.410901424
#> am            6.2223968   6.2224175   6.2224371   6.109689204   6.109545285
#> carb         -2.1186826  -2.1188479  -2.1190033  -2.994661642  -2.995048536
#> cyl           1.9782101   1.9784775   1.9787290   2.304511577   2.304864737
#> disp          0.0000000   0.0000000   0.0000000   0.005682015   0.005680537
#> drat          1.8872859   1.8874958   1.8876933   3.245911008   3.246462290
#> gear          2.2842331   2.2845511   2.2848502   3.113727570   3.114258449
#> hp            0.0000000   0.0000000   0.0000000   0.024375406   0.024380586
#> qsec          0.9018953   0.9018935   0.9018919   1.233937395   1.233937105
#> vs            2.3348126   2.3351940   2.3355527   2.977003239   2.977543415
#> wt           -2.0712938  -2.0712073  -2.0711260  -2.433881375  -2.433599910
#>                    0.0707        0.0660        0.0615        0.0574
#> (Intercept) -31.417358439 -31.423705902 -31.429941957 -31.436065028
#> am            6.109408463   6.109278366   6.109154633   6.109036918
#> carb         -2.995427849  -2.995799685  -2.996164146  -2.996521336
#> cyl           2.305209293   2.305545253   2.305872643   2.306191507
#> disp          0.005679157   0.005677869   0.005676667   0.005675545
#> drat          3.247001591   3.247529041   3.248044775   3.248548937
#> gear          3.114776949   3.115283183   3.115777276   3.116259364
#> hp            0.024385722   0.024390813   0.024395859   0.024400860
#> qsec          1.233939197   1.233943479   1.233949768   1.233957888
#> vs            2.978070313   2.978583992   2.979084537   2.979572055
#> wt           -2.433331677  -2.433075975  -2.432832127  -2.432599479
#>                    0.0535        0.0499        0.0465        0.0434
#> (Intercept) -31.442073800 -31.447967220 -31.453744475 -31.459404992
#> am            6.108924885   6.108818211   6.108716586   6.108619711
#> carb         -2.996871357  -2.997214309  -2.997550295  -2.997879415
#> cyl           2.306501903   2.306803904   2.307097598   2.307383083
#> disp          0.005674496   0.005673517   0.005672601   0.005671744
#> drat          3.249041680   3.249523158   3.249993533   3.250452974
#> gear          3.116729597   3.117188134   3.117635143   3.118070802
#> hp            0.024405816   0.024410725   0.024415586   0.024420400
#> qsec          1.233967670   1.233978951   1.233991578   1.234005403
#> vs            2.980046669   2.980508525   2.980957782   2.981394615
#> wt           -2.432377404  -2.432165298  -2.431962580  -2.431768697
#>                    0.0405        0.0377        0.0352        0.0328
#> (Intercept) -31.464948417 -31.470374613 -31.475683642 -31.480875758
#> am            6.108527300   6.108439081   6.108354792   6.108274184
#> carb         -2.998201771  -2.998517463  -2.998826591  -2.999129255
#> cyl           2.307660467   2.307929871   2.308191423   2.308445260
#> disp          0.005670941   0.005670189   0.005669481   0.005668816
#> drat          3.250901651   3.251339739   3.251767417   3.252184867
#> gear          3.118495295   3.118908815   3.119311558   3.119703729
#> hp            0.024425165   0.024429880   0.024434545   0.024439157
#> qsec          1.234020285   1.234036093   1.234052700   1.234069988
#> vs            2.981819212   2.982231777   2.982632519   2.983021661
#> wt           -2.431583117  -2.431405334  -2.431234865  -2.431071251
#>                    0.0306        0.0286        0.0266        0.0248
#> (Intercept) -31.485951393 -31.490911151 -31.495755790 -31.500486218
#> am            6.108197020   6.108123074   6.108052132   6.107983991
#> carb         -2.999425557  -2.999715597  -2.999999473  -3.000277286
#> cyl           2.308691525   2.308930369   2.309161945   2.309386415
#> disp          0.005668189   0.005667596   0.005667035   0.005666502
#> drat          3.252592274   3.252989824   3.253377705   3.253756108
#> gear          3.120085533   3.120457183   3.120818891   3.121170876
#> hp            0.024443717   0.024448224   0.024452676   0.024457073
#> qsec          1.234087844   1.234106164   1.234124850   1.234143809
#> vs            2.983399433   2.983766072   2.984121821   2.984466930
#> wt           -2.430914054  -2.430762859  -2.430617276  -2.430476932
#>                    0.0232        0.0216        0.0201        0.0188
#> (Intercept) -31.505103480 -31.509608748 -31.514003309 -31.518288560
#> am            6.107918460   6.107855357   6.107794512   6.107735764
#> carb         -3.000549136  -3.000815122  -3.001075342  -3.001329897
#> cyl           2.309603943   2.309814694   2.310018838   2.310216547
#> disp          0.005665994   0.005665509   0.005665045   0.005664598
#> drat          3.254125222   3.254485239   3.254836350   3.255178747
#> gear          3.121513354   3.121846546   3.122170671   3.122485950
#> hp            0.024461413   0.024465696   0.024469921   0.024474087
#> qsec          1.234162955   1.234182209   1.234201497   1.234220751
#> vs            2.984801649   2.985126236   2.985440949   2.985746046
#> wt           -2.430341476  -2.430210580  -2.430083931  -2.429961238
#>                    0.0175        0.0163        0.0152        0.0142
#> (Intercept) -31.522465994 -31.526537194 -31.530503821 -31.534367610
#> am            6.107678962   6.107623965   6.107570639   6.107518862
#> carb         -3.001578884  -3.001822401  -3.002060547  -3.002293419
#> cyl           2.310407993   2.310593350   2.310772790   2.310946487
#> disp          0.005664168   0.005663751   0.005663347   0.005662952
#> drat          3.255512621   3.255838162   3.256155559   3.256465000
#> gear          3.122792603   3.123090848   3.123380904   3.123662986
#> hp            0.024478194   0.024482241   0.024486226   0.024490151
#> qsec          1.234239908   1.234258912   1.234277709   1.234296252
#> vs            2.986041788   2.986328436   2.986606249   2.986875485
#> wt           -2.429842228  -2.429726645  -2.429614249  -2.429504817
#>                    0.0133       0.0124        0.0115        0.0107
#> (Intercept) -31.538130357 -31.54179392 -31.545360189 -31.548831115
#> am            6.107468519   6.10741950   6.107371710   6.107325054
#> carb         -3.002521113  -3.00274373  -3.002961353  -3.003174090
#> cyl           2.311114613   2.31127734   2.311434839   2.311587276
#> disp          0.005662567   0.00566219   0.005661819   0.005661453
#> drat          3.256766673   3.25706076   3.257347450   3.257626918
#> gear          3.123937310   3.12420409   3.124463528   3.124715838
#> hp            0.024494014   0.02449781   0.024501552   0.024505227
#> qsec          1.234314500   1.23433241   1.234349956   1.234367101
#> vs            2.987136402   2.98738925   2.987634288   2.987871757
#> wt           -2.429398143  -2.42929403  -2.429192312  -2.429092812
#>                    0.0100        0.0093        0.0087        0.0081
#> (Intercept) -31.552208671 -31.555494861 -31.558691710 -31.561801258
#> am            6.107279447   6.107234812   6.107191075   6.107148172
#> carb         -3.003382030  -3.003585268  -3.003783897  -3.003978008
#> cyl           2.311734818   2.311877628   2.312015867   2.312149694
#> disp          0.005661091   0.005660733   0.005660377   0.005660023
#> drat          3.257899345   3.258164907   3.258423779   3.258676130
#> gear          3.124961222   3.125199879   3.125432007   3.125657798
#> hp            0.024508839   0.024512388   0.024515874   0.024519296
#> qsec          1.234383821   1.234400093   1.234415899   1.234431221
#> vs            2.988101904   2.988324969   2.988541189   2.988750794
#> wt           -2.428995382  -2.428899882  -2.428806186  -2.428714175
#>                   0.0076        0.0071        0.0066        0.0062
#> (Intercept) -31.56482556 -31.567766663 -31.570626632 -31.573407517
#> am            6.10710604   6.107064631   6.107023889   6.106983770
#> carb         -3.00416769  -3.004353041  -3.004534142  -3.004711085
#> cyl           2.31227926   2.312404724   2.312526228   2.312643918
#> disp          0.00565967   0.005659318   0.005658966   0.005658614
#> drat          3.25892213   3.259161945   3.259395735   3.259623661
#> gear          3.12587744   3.126091124   3.126299026   3.126501325
#> hp            0.02452266   0.024525951   0.024529184   0.024532353
#> qsec          1.23444605   1.234460366   1.234474171   1.234487457
#> vs            2.98895401   2.989151062   2.989342163   2.989527526
#> wt           -2.42862374  -2.428534793  -2.428447238  -2.428360998
#> 
#> $iter
#>   [1]   0   4   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   4   4   4   4   4   4  41  53  88  13  14  13  13  13
#>  [37]  13  13  12   7   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [55]   1 244   1   1   1   1   1   1 148   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 6.151405893 5.736821724 5.350179140 4.989594972 4.653312971 4.339695251
#>   [7] 4.047214315 3.774445616 3.520060614 3.282820310 3.061569208 2.855229690
#>  [13] 2.662796765 2.483333175 2.315964831 2.159876552 2.014308102 1.878550477
#>  [19] 1.751942462 1.633867404 1.523750209 1.421054545 1.325280225 1.235960773
#>  [25] 1.152661154 1.074975650 1.002525889 0.934958999 0.871945892 0.813179657
#>  [31] 0.758374070 0.707262196 0.659595091 0.615140589 0.573682172 0.535017914
#>  [37] 0.498959497 0.465331297 0.433969525 0.404721432 0.377444562 0.352006061
#>  [43] 0.328282031 0.306156921 0.285522969 0.266279677 0.248333318 0.231596483
#>  [49] 0.215987655 0.201430810 0.187855048 0.175194246 0.163386740 0.152375021
#>  [55] 0.142105455 0.132528022 0.123596077 0.115266115 0.107497565 0.100252589
#>  [61] 0.093495900 0.087194589 0.081317966 0.075837407 0.070726220 0.065959509
#>  [67] 0.061514059 0.057368217 0.053501791 0.049895950 0.046533130 0.043396953
#>  [73] 0.040472143 0.037744456 0.035200606 0.032828203 0.030615692 0.028552297
#>  [79] 0.026627968 0.024833332 0.023159648 0.021598766 0.020143081 0.018785505
#>  [85] 0.017519425 0.016338674 0.015237502 0.014210545 0.013252802 0.012359608
#>  [91] 0.011526612 0.010749756 0.010025259 0.009349590 0.008719459 0.008131797
#>  [97] 0.007583741 0.007072622 0.006595951 0.006151406
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 28
#> 
#> $loss
#>   [1] 999.88286 847.33683 719.70940 613.41458 525.35641 452.86356 393.63261
#>   [8] 345.67874 307.29303 277.00531 253.55200 235.84818 222.96336 214.10049
#>  [15] 208.57774 205.81276 205.24717 205.24717 205.24717 205.24717 205.24717
#>  [22] 205.24717 201.50355 193.84664 187.48267 182.22143 177.90085 174.38120
#>  [29] 160.36911 122.09175 104.02633  99.64306  94.79331  91.08531  88.32736
#>  [36]  86.36820  85.07418  84.33103  84.04141  84.02960  84.02960  84.02959
#>  [43]  84.02959  84.02959  84.02959  84.02959  84.02959  84.02959  84.02959
#>  [50]  84.02959  84.02959  84.02959  84.02959  84.02959  84.02959  70.70337
#>  [57]  70.70336  70.70335  70.70333  70.70332  70.70331  70.70331  66.38256
#>  [64]  66.38253  66.38250  66.38248  66.38245  66.38243  66.38240  66.38238
#>  [71]  66.38236  66.38234  66.38232  66.38230  66.38229  66.38227  66.38225
#>  [78]  66.38224  66.38223  66.38221  66.38220  66.38219  66.38218  66.38216
#>  [85]  66.38215  66.38214  66.38213  66.38213  66.38212  66.38211  66.38210
#>  [92]  66.38209  66.38209  66.38208  66.38207  66.38207  66.38206  66.38206
#>  [99]  66.38205  66.38205
#> 
#> $linear.predictors
#>      6.1514   5.7368   5.3502   4.9896   4.6533   4.3397   4.0472   3.7744
#> 1  19.97143 20.04325 20.11023 20.17270 20.23095 20.28528 20.33595 20.38320
#> 2  19.97143 19.84472 19.72656 19.61636 19.51358 19.41773 19.32834 19.24498
#> 3  19.97143 20.07960 20.18048 20.27456 20.36230 20.44413 20.52044 20.59161
#> 4  19.97143 19.91742 19.86706 19.82009 19.77628 19.73543 19.69733 19.66180
#> 5  19.97143 19.56511 19.18617 18.83278 18.50320 18.19583 17.90918 17.64185
#> 6  19.97143 19.75525 19.55363 19.36561 19.19026 19.02672 18.87421 18.73198
#> 7  19.97143 19.72728 19.49960 19.28725 19.08922 18.90453 18.73230 18.57167
#> 8  19.97143 18.90522 17.91087 16.98353 16.11869 15.31214 14.55995 13.85846
#> 9  19.97143 18.80791 17.72281 16.71084 15.76708 14.88692 14.06608 13.30057
#> 10 19.97143 18.85209 17.80819 16.83465 15.92672 15.07998 14.29031 13.55386
#> 11 19.97143 20.61087 21.20721 21.76336 22.28203 22.76574 23.21685 23.63756
#> 12 19.97143 20.93802 21.83946 22.68015 23.46418 24.19537 24.87728 25.51323
#> 13 19.97143 20.81499 21.60169 22.33537 23.01961 23.65773 24.25284 24.80785
#> 14 19.97143 20.46267 20.92081 21.34807 21.74653 22.11813 22.46470 22.78790
#> 15 19.97143 19.92022 19.87246 19.82792 19.78638 19.74765 19.71152 19.67783
#> 16 19.97143 19.69373 19.43475 19.19322 18.96797 18.75791 18.56200 18.37929
#> 17 19.97143 19.69094 19.42935 19.18539 18.95787 18.74569 18.54780 18.36326
#> 18 19.97143 20.64442 21.27206 21.85739 22.40328 22.91237 23.38715 23.82994
#> 19 19.97143 20.99506 21.94970 22.84000 23.67030 24.44464 25.16679 25.84027
#> 20 19.97143 19.84472 19.72656 19.61636 19.51358 19.41773 19.32834 19.24498
#> 21 19.97143 20.28651 20.58036 20.85441 21.10999 21.34834 21.57062 21.77793
#>      3.5201   3.2828   3.0616   2.8552   2.6628    2.4833    2.3160    2.1599
#> 1  20.42727 20.46837 20.50670 20.54245 20.57578 20.606872 20.635866 20.662907
#> 2  19.16724 19.09473 19.02711 18.96405 18.90524 18.850390 18.799239 18.751535
#> 3  20.65798 20.71988 20.77761 20.83145 20.88166 20.928481 20.972150 21.012876
#> 4  19.62866 19.59775 19.56893 19.54205 19.51699 19.493609 19.471807 19.451474
#> 5  17.39253 17.16002 16.94318 16.74096 16.55236 16.376472 16.212440 16.059463
#> 6  18.59933 18.47562 18.36025 18.25266 18.15232 18.058737 17.971463 17.890072
#> 7  18.42186 18.28215 18.15186 18.03035 17.91703 17.811345 17.712784 17.620865
#> 8  13.20424 12.59411 12.02511 11.49445 10.99956 10.538026 10.107594  9.706173
#> 9  12.58664 11.92084 11.29990 10.72082 10.18076  9.677102  9.207388  8.769332
#> 10 12.86705 12.22652 11.62916 11.07207 10.55252 10.067981  9.616103  9.194679
#> 11 24.02992 24.39583 24.73708 25.05533 25.35213 25.628926 25.887068 26.127813
#> 12 26.10632 26.65943 27.17527 27.65635 28.10500 28.523410 28.913623 29.277537
#> 13 25.32545 25.80816 26.25834 26.67818 27.06973 27.434886 27.775432 28.093026
#> 14 23.08932 23.37043 23.63259 23.87709 24.10510 24.317749 24.516065 24.701015
#> 15 19.64640 19.61710 19.58977 19.56428 19.54052 19.518348 19.497675 19.478395
#> 16 18.20890 18.04999 17.90179 17.76358 17.63468 17.514475 17.402368 17.297816
#> 17 18.19115 18.03064 17.88095 17.74135 17.61115 17.489735 17.376500 17.270896
#> 18 24.24288 24.62799 24.98715 25.32210 25.63447 25.925796 26.197484 26.450862
#> 19 26.46836 27.05411 27.60039 28.10986 28.58498 29.028089 29.441330 29.826720
#> 20 19.16724 19.09473 19.02711 18.96405 18.90524 18.850390 18.799239 18.751535
#> 21 21.97126 22.15157 22.31972 22.47654 22.62279 22.759181 22.886381 23.005009
#>       2.0143    1.8786    1.7519    1.6339    1.5238    1.4211    1.3253
#> 1  20.681860 20.681860 20.681860 20.681860 20.681860 20.681860 20.734858
#> 2  18.718098 18.718098 18.718098 18.718098 18.718098 18.718098 18.640190
#> 3  21.041422 21.041422 21.041422 21.041422 21.041422 21.041422 21.224366
#> 4  19.437222 19.437222 19.437222 19.437222 19.437222 19.437222 19.473153
#> 5  15.952236 15.952236 15.952236 15.952236 15.952236 15.952236 15.940836
#> 6  17.833022 17.833022 17.833022 17.833022 17.833022 17.833022 17.824103
#> 7  17.556436 17.556436 17.556436 17.556436 17.556436 17.556436 17.563400
#> 8   9.424802  9.424802  9.424802  9.424802  9.424802  9.424802  9.453002
#> 9   8.462282  8.462282  8.462282  8.462282  8.462282  8.462282  8.487031
#> 10  8.899288  8.899288  8.899288  8.899288  8.899288  8.899288  8.907720
#> 11 26.296560 26.296560 26.296560 26.296560 26.296560 26.296560 26.335507
#> 12 29.532618 29.532618 29.532618 29.532618 29.532618 29.532618 29.526867
#> 13 28.315639 28.315639 28.315639 28.315639 28.315639 28.315639 28.365396
#> 14 24.830653 24.830653 24.830653 24.830653 24.830653 24.830653 24.893997
#> 15 19.464881 19.464881 19.464881 19.464881 19.464881 19.464881 19.440197
#> 16 17.224533 17.224533 17.224533 17.224533 17.224533 17.224533 17.134394
#> 17 17.196874 17.196874 17.196874 17.196874 17.196874 17.196874 17.168864
#> 18 26.628463 26.628463 26.628463 26.628463 26.628463 26.628463 26.561702
#> 19 30.096854 30.096854 30.096854 30.096854 30.096854 30.096854 30.028280
#> 20 18.718098 18.718098 18.718098 18.718098 18.718098 18.718098 18.593271
#> 21 23.088160 23.088160 23.088160 23.088160 23.088160 23.088160 23.102866
#>       1.2360    1.1527    1.0750    1.0025    0.9350    0.8719    0.8132
#> 1  20.852096 20.961402 21.063341 21.158410 21.247071 21.214431 20.749384
#> 2  18.467847 18.307165 18.157312 18.017558 17.887224 17.509488 16.188747
#> 3  21.629066 22.006383 22.358270 22.686442 22.992495 23.347393 23.943506
#> 4  19.552639 19.626746 19.695859 19.760314 19.820424 19.770758 19.325817
#> 5  15.915617 15.892105 15.870177 15.849727 15.830655 15.734706 15.349938
#> 6  17.804372 17.785976 17.768820 17.752821 17.737899 17.596342 16.982266
#> 7  17.578805 17.593168 17.606563 17.619055 17.630705 17.545583 17.103899
#> 8   9.515383  9.573544  9.627786  9.678371  9.725547  9.935436 10.757957
#> 9   8.541779  8.592822  8.640426  8.684821  8.726224  8.952778  9.874035
#> 10  8.926372  8.943762  8.959980  8.975106  8.989211  9.153464  9.877261
#> 11 26.421665 26.501993 26.576907 26.646772 26.711928 27.001707 28.137223
#> 12 29.514146 29.502285 29.491224 29.480908 29.471288 29.532728 29.855786
#> 13 28.475466 28.578088 28.673794 28.763050 28.846290 29.112039 30.070911
#> 14 25.034125 25.164771 25.286612 25.400242 25.506213 25.387884 24.458611
#> 15 19.385592 19.334681 19.287202 19.242923 19.201628 18.965662 17.999709
#> 16 16.934994 16.749086 16.575707 16.414014 16.263218 15.890297 14.664082
#> 17 17.106903 17.049134 16.995258 16.945014 16.898156 16.716166 16.023882
#> 18 26.414016 26.276323 26.147910 26.028152 25.916465 25.873328 26.062458
#> 19 29.876584 29.735152 29.603252 29.480242 29.365522 29.218726 28.930466
#> 20 18.317136 18.059685 17.819585 17.595667 17.386840 17.389748 18.136809
#> 21 23.135397 23.165727 23.194013 23.220393 23.244995 23.551336 24.907251
#>      0.7584   0.7073   0.6596   0.6151   0.5737   0.5350   0.4990   0.4653
#> 1  21.30173 21.30052 21.29859 21.29691 21.29524 21.29368 21.29223 21.29088
#> 2  15.44507 15.32002 15.16221 15.01485 14.87767 14.74973 14.63042 14.51915
#> 3  19.79993 20.15030 20.60324 21.02498 21.41890 21.78628 22.12890 22.44842
#> 4  15.76446 15.89737 16.07240 16.23499 16.38725 16.52927 16.66171 16.78522
#> 5  15.87816 15.82503 15.75671 15.69308 15.63367 15.57827 15.52660 15.47841
#> 6  16.71348 16.67706 16.63134 16.58865 16.54891 16.51186 16.47731 16.44508
#> 7  16.59064 16.58797 16.58559 16.58327 16.58122 16.57932 16.57754 16.57588
#> 8  11.31762 11.37379 11.44420 11.50999 11.57118 11.62825 11.68147 11.73111
#> 9  10.89013 10.93290 10.98547 11.03471 11.08039 11.12298 11.16271 11.19975
#> 10 11.08422 11.09322 11.10253 11.11145 11.11951 11.12703 11.13403 11.14057
#> 11 29.77062 29.81718 29.87360 29.92656 29.97558 30.02128 30.06391 30.10366
#> 12 29.54636 29.57239 29.60584 29.63697 29.66604 29.69316 29.71844 29.74203
#> 13 30.66736 30.75001 30.85410 30.95131 31.04179 31.12617 31.20487 31.27826
#> 14 23.14434 23.19085 23.25279 23.31032 23.36423 23.41450 23.46139 23.50512
#> 15 19.09973 18.97803 18.82174 18.67616 18.54026 18.41352 18.29532 18.18509
#> 16 14.78173 14.62054 14.41532 14.22389 14.04546 13.87905 13.72386 13.57913
#> 17 18.09244 17.94984 17.76533 17.59361 17.43316 17.28351 17.14395 17.01379
#> 18 28.25653 28.12934 27.96367 27.80954 27.66544 27.53104 27.40570 27.28880
#> 19 29.79696 29.68634 29.54401 29.41140 29.28763 29.17220 29.06454 28.96415
#> 20 14.77432 14.82980 14.90470 14.97388 15.03902 15.09978 15.15644 15.20929
#> 21 26.68417 26.71749 26.75663 26.79348 26.82745 26.85912 26.88866 26.91621
#>      0.4340   0.4047   0.3774   0.3520   0.3283   0.3062   0.2855   0.2663
#> 1  21.28972 21.28923 21.28916 21.28913 21.28912 21.28911 21.28910 21.28910
#> 2  14.41516 14.38995 14.39026 14.39049 14.39065 14.39077 14.39086 14.39092
#> 3  22.74578 22.81891 22.81917 22.81934 22.81945 22.81952 22.81958 22.81961
#> 4  16.89977 16.92832 16.92876 16.92907 16.92928 16.92943 16.92953 16.92960
#> 5  15.43355 15.42249 15.42248 15.42248 15.42248 15.42249 15.42249 15.42250
#> 6  16.41495 16.40762 16.40773 16.40782 16.40788 16.40793 16.40796 16.40798
#> 7  16.57422 16.57388 16.57398 16.57406 16.57412 16.57416 16.57419 16.57421
#> 8  11.77754 11.78878 11.78859 11.78846 11.78835 11.78828 11.78823 11.78819
#> 9  11.23453 11.24285 11.24261 11.24243 11.24229 11.24220 11.24213 11.24208
#> 10 11.14691 11.14826 11.14803 11.14786 11.14774 11.14765 11.14759 11.14755
#> 11 30.14108 30.14993 30.14957 30.14931 30.14913 30.14899 30.14890 30.14883
#> 12 29.76397 29.76942 29.76942 29.76942 29.76942 29.76941 29.76941 29.76940
#> 13 31.34686 31.36349 31.36327 31.36310 31.36298 31.36289 31.36283 31.36278
#> 14 23.54567 23.55565 23.55586 23.55601 23.55613 23.55621 23.55627 23.55631
#> 15 18.08244 18.05715 18.05715 18.05717 18.05719 18.05721 18.05722 18.05723
#> 16 13.44409 13.41116 13.41136 13.41152 13.41164 13.41173 13.41179 13.41183
#> 17 16.89275 16.86280 16.86266 16.86257 16.86252 16.86249 16.86247 16.86246
#> 18 27.18016 27.15334 27.15312 27.15296 27.15286 27.15278 27.15273 27.15269
#> 19 28.87063 28.84772 28.84771 28.84771 28.84771 28.84771 28.84771 28.84771
#> 20 15.25792 15.27067 15.27110 15.27137 15.27154 15.27166 15.27174 15.27179
#> 21 26.94228 26.94838 26.94800 26.94772 26.94752 26.94738 26.94727 26.94720
#>      0.2483   0.2316   0.2160   0.2014   0.1879   0.1752   0.1634   0.1524
#> 1  21.28910 21.28910 21.28910 21.28910 21.28910 21.28910 21.28910 21.28910
#> 2  14.39096 14.39099 14.39101 14.39103 14.39104 14.39105 14.39105 14.39106
#> 3  22.81964 22.81965 22.81967 22.81967 22.81968 22.81968 22.81969 22.81969
#> 4  16.92965 16.92969 16.92971 16.92973 16.92974 16.92975 16.92976 16.92976
#> 5  15.42250 15.42250 15.42250 15.42250 15.42250 15.42250 15.42250 15.42250
#> 6  16.40800 16.40801 16.40802 16.40803 16.40803 16.40803 16.40803 16.40804
#> 7  16.57423 16.57424 16.57424 16.57425 16.57425 16.57426 16.57426 16.57426
#> 8  11.78817 11.78815 11.78813 11.78812 11.78812 11.78811 11.78811 11.78811
#> 9  11.24205 11.24203 11.24201 11.24200 11.24199 11.24198 11.24198 11.24198
#> 10 11.14752 11.14750 11.14748 11.14747 11.14746 11.14746 11.14745 11.14745
#> 11 30.14878 30.14875 30.14873 30.14871 30.14870 30.14869 30.14868 30.14868
#> 12 29.76940 29.76940 29.76940 29.76940 29.76940 29.76940 29.76940 29.76940
#> 13 31.36275 31.36273 31.36271 31.36270 31.36269 31.36269 31.36269 31.36268
#> 14 23.55634 23.55637 23.55638 23.55639 23.55640 23.55641 23.55641 23.55641
#> 15 18.05724 18.05725 18.05725 18.05725 18.05726 18.05726 18.05726 18.05726
#> 16 13.41186 13.41188 13.41190 13.41191 13.41192 13.41192 13.41193 13.41193
#> 17 16.86245 16.86244 16.86244 16.86244 16.86243 16.86243 16.86243 16.86243
#> 18 27.15267 27.15265 27.15263 27.15263 27.15262 27.15261 27.15261 27.15261
#> 19 28.84771 28.84771 28.84772 28.84772 28.84772 28.84772 28.84772 28.84772
#> 20 15.27183 15.27185 15.27187 15.27188 15.27189 15.27190 15.27190 15.27190
#> 21 26.94715 26.94711 26.94709 26.94707 26.94706 26.94705 26.94704 26.94704
#>      0.1421   0.1325   0.1236   0.1153   0.1075   0.1003   0.0935   0.0872
#> 1  21.28910 21.82020 21.82037 21.82053 21.82068 21.82082 21.82095 21.82108
#> 2  14.39106 13.35027 13.35009 13.34991 13.34975 13.34959 13.34945 13.34931
#> 3  22.81969 22.87086 22.87075 22.87065 22.87055 22.87046 22.87037 22.87030
#> 4  16.92977 18.38055 18.38074 18.38093 18.38110 18.38126 18.38141 18.38155
#> 5  15.42250 15.57522 15.57526 15.57530 15.57533 15.57536 15.57540 15.57542
#> 6  16.40804 16.45998 16.45998 16.45998 16.45998 16.45998 16.45998 16.45999
#> 7  16.57426 16.71715 16.71716 16.71717 16.71717 16.71718 16.71718 16.71719
#> 8  11.78811 11.27173 11.27165 11.27157 11.27150 11.27143 11.27137 11.27131
#> 9  11.24197 10.89899 10.89895 10.89891 10.89887 10.89884 10.89880 10.89877
#> 10 11.14745 11.13574 11.13575 11.13576 11.13578 11.13579 11.13580 11.13580
#> 11 30.14868 30.38746 30.38753 30.38760 30.38767 30.38773 30.38779 30.38785
#> 12 29.76940 30.22804 30.22807 30.22810 30.22813 30.22815 30.22817 30.22819
#> 13 31.36268 31.79553 31.79560 31.79567 31.79573 31.79579 31.79585 31.79590
#> 14 23.55641 21.10364 21.10321 21.10281 21.10242 21.10207 21.10173 21.10141
#> 15 18.05726 19.06937 19.06957 19.06976 19.06994 19.07011 19.07027 19.07042
#> 16 13.41193 13.38397 13.38396 13.38394 13.38393 13.38392 13.38390 13.38389
#> 17 16.86243 17.86246 17.86269 17.86290 17.86311 17.86330 17.86348 17.86365
#> 18 27.15261 26.50550 26.50538 26.50526 26.50515 26.50505 26.50495 26.50486
#> 19 28.84772 29.07290 29.07302 29.07313 29.07324 29.07333 29.07342 29.07351
#> 20 15.27191 15.17055 15.17045 15.17035 15.17026 15.17018 15.17010 15.17003
#> 21 26.94703 26.33990 26.33983 26.33978 26.33972 26.33967 26.33962 26.33958
#>      0.0813   0.0758   0.0707   0.0660   0.0615   0.0574   0.0535   0.0499
#> 1  22.05373 22.05382 22.05390 22.05399 22.05408 22.05416 22.05425 22.05433
#> 2  13.68778 13.68759 13.68741 13.68724 13.68708 13.68692 13.68677 13.68663
#> 3  23.68640 23.68647 23.68653 23.68659 23.68666 23.68673 23.68679 23.68686
#> 4  17.49932 17.49950 17.49967 17.49983 17.49997 17.50011 17.50023 17.50034
#> 5  14.87319 14.87324 14.87328 14.87332 14.87335 14.87338 14.87341 14.87343
#> 6  15.94749 15.94745 15.94740 15.94736 15.94731 15.94726 15.94721 15.94716
#> 7  16.31937 16.31934 16.31931 16.31928 16.31925 16.31921 16.31918 16.31914
#> 8  10.99200 10.99176 10.99152 10.99129 10.99107 10.99086 10.99065 10.99045
#> 9  10.77386 10.77377 10.77369 10.77361 10.77353 10.77345 10.77338 10.77330
#> 10 11.47111 11.47124 11.47136 11.47148 11.47160 11.47171 11.47182 11.47193
#> 11 30.33014 30.33021 30.33028 30.33034 30.33041 30.33047 30.33053 30.33059
#> 12 29.98778 29.98770 29.98762 29.98754 29.98746 29.98739 29.98731 29.98723
#> 13 32.13597 32.13602 32.13607 32.13612 32.13618 32.13623 32.13629 32.13634
#> 14 20.88550 20.88515 20.88480 20.88446 20.88414 20.88382 20.88351 20.88320
#> 15 18.97861 18.97872 18.97882 18.97892 18.97902 18.97912 18.97921 18.97930
#> 16 14.13109 14.13128 14.13147 14.13165 14.13183 14.13201 14.13218 14.13235
#> 17 18.59988 18.60005 18.60022 18.60039 18.60057 18.60074 18.60091 18.60108
#> 18 26.18206 26.18197 26.18189 26.18181 26.18173 26.18164 26.18156 26.18148
#> 19 29.18266 29.18273 29.18280 29.18287 29.18293 29.18300 29.18307 29.18313
#> 20 15.44589 15.44581 15.44573 15.44566 15.44559 15.44553 15.44546 15.44540
#> 21 26.23617 26.23619 26.23621 26.23623 26.23625 26.23627 26.23628 26.23629
#>      0.0465   0.0434   0.0405   0.0377   0.0352   0.0328   0.0306   0.0286
#> 1  22.05442 22.05450 22.05458 22.05466 22.05474 22.05482 22.05489 22.05497
#> 2  13.68649 13.68637 13.68624 13.68612 13.68601 13.68591 13.68580 13.68571
#> 3  23.68693 23.68699 23.68706 23.68713 23.68720 23.68727 23.68733 23.68740
#> 4  17.50045 17.50054 17.50062 17.50070 17.50077 17.50083 17.50088 17.50093
#> 5  14.87345 14.87346 14.87347 14.87348 14.87348 14.87348 14.87348 14.87348
#> 6  15.94711 15.94706 15.94701 15.94696 15.94691 15.94686 15.94681 15.94676
#> 7  16.31911 16.31908 16.31904 16.31901 16.31897 16.31894 16.31890 16.31887
#> 8  10.99025 10.99006 10.98987 10.98969 10.98952 10.98934 10.98918 10.98902
#> 9  10.77323 10.77316 10.77310 10.77303 10.77297 10.77291 10.77285 10.77279
#> 10 11.47204 11.47215 11.47225 11.47235 11.47245 11.47255 11.47264 11.47273
#> 11 30.33064 30.33070 30.33075 30.33081 30.33086 30.33091 30.33096 30.33100
#> 12 29.98716 29.98708 29.98701 29.98694 29.98686 29.98679 29.98672 29.98665
#> 13 32.13640 32.13646 32.13651 32.13657 32.13663 32.13668 32.13674 32.13679
#> 14 20.88291 20.88262 20.88235 20.88208 20.88182 20.88157 20.88132 20.88108
#> 15 18.97939 18.97948 18.97956 18.97964 18.97972 18.97980 18.97987 18.97994
#> 16 14.13252 14.13268 14.13285 14.13300 14.13316 14.13331 14.13346 14.13361
#> 17 18.60125 18.60142 18.60158 18.60175 18.60191 18.60207 18.60223 18.60239
#> 18 26.18140 26.18132 26.18124 26.18117 26.18109 26.18102 26.18094 26.18087
#> 19 29.18319 29.18326 29.18332 29.18338 29.18344 29.18350 29.18356 29.18361
#> 20 15.44535 15.44530 15.44524 15.44520 15.44515 15.44511 15.44507 15.44503
#> 21 26.23630 26.23632 26.23633 26.23634 26.23635 26.23636 26.23637 26.23637
#>      0.0266   0.0248   0.0232   0.0216   0.0201   0.0188   0.0175   0.0163
#> 1  22.05504 22.05511 22.05518 22.05525 22.05531 22.05538 22.05544 22.05550
#> 2  13.68561 13.68552 13.68544 13.68536 13.68528 13.68521 13.68514 13.68507
#> 3  23.68747 23.68754 23.68760 23.68767 23.68773 23.68780 23.68786 23.68793
#> 4  17.50097 17.50100 17.50103 17.50105 17.50107 17.50109 17.50110 17.50110
#> 5  14.87348 14.87347 14.87347 14.87346 14.87345 14.87344 14.87343 14.87342
#> 6  15.94671 15.94666 15.94661 15.94657 15.94652 15.94647 15.94643 15.94638
#> 7  16.31883 16.31880 16.31876 16.31873 16.31870 16.31866 16.31863 16.31860
#> 8  10.98886 10.98870 10.98855 10.98841 10.98827 10.98813 10.98799 10.98786
#> 9  10.77273 10.77267 10.77262 10.77257 10.77252 10.77247 10.77242 10.77237
#> 10 11.47282 11.47291 11.47300 11.47309 11.47317 11.47326 11.47334 11.47342
#> 11 30.33105 30.33109 30.33114 30.33118 30.33122 30.33126 30.33130 30.33134
#> 12 29.98657 29.98650 29.98643 29.98636 29.98629 29.98622 29.98615 29.98608
#> 13 32.13685 32.13690 32.13696 32.13701 32.13706 32.13711 32.13716 32.13721
#> 14 20.88086 20.88063 20.88042 20.88021 20.88001 20.87982 20.87963 20.87945
#> 15 18.98001 18.98007 18.98013 18.98019 18.98025 18.98031 18.98036 18.98041
#> 16 14.13376 14.13390 14.13404 14.13418 14.13432 14.13445 14.13458 14.13471
#> 17 18.60254 18.60269 18.60284 18.60298 18.60312 18.60326 18.60339 18.60353
#> 18 26.18080 26.18074 26.18067 26.18061 26.18055 26.18049 26.18043 26.18037
#> 19 29.18367 29.18372 29.18378 29.18383 29.18388 29.18393 29.18398 29.18402
#> 20 15.44499 15.44496 15.44492 15.44489 15.44486 15.44483 15.44480 15.44477
#> 21 26.23638 26.23639 26.23640 26.23641 26.23642 26.23643 26.23644 26.23646
#>      0.0152   0.0142   0.0133   0.0124   0.0115   0.0107   0.0100   0.0093
#> 1  22.05556 22.05561 22.05567 22.05572 22.05577 22.05582 22.05587 22.05591
#> 2  13.68500 13.68494 13.68488 13.68483 13.68477 13.68472 13.68467 13.68462
#> 3  23.68799 23.68805 23.68812 23.68818 23.68824 23.68830 23.68835 23.68841
#> 4  17.50111 17.50111 17.50111 17.50110 17.50110 17.50109 17.50108 17.50107
#> 5  14.87341 14.87340 14.87339 14.87338 14.87337 14.87336 14.87335 14.87334
#> 6  15.94634 15.94629 15.94625 15.94620 15.94616 15.94612 15.94608 15.94604
#> 7  16.31857 16.31853 16.31850 16.31847 16.31844 16.31841 16.31838 16.31836
#> 8  10.98773 10.98760 10.98748 10.98736 10.98724 10.98713 10.98701 10.98690
#> 9  10.77232 10.77228 10.77224 10.77219 10.77215 10.77211 10.77207 10.77204
#> 10 11.47350 11.47357 11.47365 11.47372 11.47380 11.47387 11.47394 11.47401
#> 11 30.33137 30.33141 30.33145 30.33148 30.33151 30.33155 30.33158 30.33161
#> 12 29.98601 29.98594 29.98587 29.98580 29.98574 29.98567 29.98560 29.98554
#> 13 32.13726 32.13730 32.13735 32.13739 32.13744 32.13748 32.13752 32.13756
#> 14 20.87928 20.87911 20.87895 20.87880 20.87865 20.87850 20.87836 20.87823
#> 15 18.98046 18.98051 18.98055 18.98059 18.98064 18.98067 18.98071 18.98075
#> 16 14.13484 14.13496 14.13508 14.13520 14.13532 14.13544 14.13555 14.13566
#> 17 18.60366 18.60378 18.60390 18.60402 18.60414 18.60425 18.60436 18.60447
#> 18 26.18032 26.18027 26.18022 26.18017 26.18013 26.18008 26.18004 26.18000
#> 19 29.18407 29.18411 29.18416 29.18420 29.18424 29.18428 29.18432 29.18436
#> 20 15.44474 15.44472 15.44469 15.44467 15.44465 15.44462 15.44460 15.44458
#> 21 26.23647 26.23648 26.23649 26.23650 26.23652 26.23653 26.23655 26.23656
#>      0.0087   0.0081   0.0076   0.0071   0.0066   0.0062
#> 1  22.05596 22.05600 22.05604 22.05608 22.05612 22.05615
#> 2  13.68457 13.68452 13.68448 13.68444 13.68440 13.68436
#> 3  23.68847 23.68852 23.68858 23.68863 23.68868 23.68874
#> 4  17.50106 17.50104 17.50103 17.50101 17.50100 17.50098
#> 5  14.87333 14.87332 14.87331 14.87330 14.87329 14.87328
#> 6  15.94600 15.94596 15.94593 15.94589 15.94586 15.94582
#> 7  16.31833 16.31830 16.31827 16.31825 16.31822 16.31820
#> 8  10.98679 10.98669 10.98659 10.98648 10.98639 10.98629
#> 9  10.77200 10.77196 10.77193 10.77189 10.77186 10.77183
#> 10 11.47408 11.47414 11.47421 11.47428 11.47434 11.47440
#> 11 30.33164 30.33167 30.33170 30.33173 30.33176 30.33178
#> 12 29.98547 29.98540 29.98534 29.98528 29.98521 29.98515
#> 13 32.13760 32.13764 32.13767 32.13771 32.13774 32.13777
#> 14 20.87810 20.87797 20.87785 20.87774 20.87763 20.87752
#> 15 18.98078 18.98081 18.98084 18.98087 18.98090 18.98092
#> 16 14.13577 14.13588 14.13598 14.13609 14.13619 14.13629
#> 17 18.60457 18.60467 18.60477 18.60486 18.60495 18.60504
#> 18 26.17996 26.17992 26.17989 26.17985 26.17982 26.17979
#> 19 29.18439 29.18443 29.18446 29.18450 29.18453 29.18456
#> 20 15.44456 15.44454 15.44452 15.44450 15.44448 15.44446
#> 21 26.23658 26.23659 26.23661 26.23662 26.23664 26.23666
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.4 14.3 22.8 17.8 16.4 17.3 15.2 10.4 10.4 14.7 32.4 30.4 33.9 21.5 15.2
#> [16] 13.3 19.2 26.0 30.4 15.0 21.4
#> 
#> $X
#>               am       carb        cyl        disp        drat       gear
#>  [1,] -0.7071068 -1.1364823 -0.1519109  0.09204202 -0.91994160 -0.7844645
#>  [2,] -0.7071068  0.7476858  0.9114654  0.86458908 -0.68211579 -0.7844645
#>  [3,] -0.7071068 -0.5084263 -1.2152872 -0.79562970  0.61677903  0.5883484
#>  [4,] -0.7071068  0.7476858 -0.1519109 -0.59264674  0.61677903  0.5883484
#>  [5,] -0.7071068  0.1196297  0.9114654  0.22685906 -0.93823590 -0.7844645
#>  [6,] -0.7071068  0.1196297  0.9114654  0.22685906 -0.93823590 -0.7844645
#>  [7,] -0.7071068  0.1196297  0.9114654  0.22685906 -0.93823590 -0.7844645
#>  [8,] -0.7071068  0.7476858  0.9114654  1.71287604 -1.19435600 -0.7844645
#>  [9,] -0.7071068  0.7476858  0.9114654  1.62198815 -1.06629595 -0.7844645
#> [10,] -0.7071068  0.7476858  0.9114654  1.47050834 -0.64552720 -0.7844645
#> [11,]  1.4142136 -1.1364823 -1.2152872 -1.26597452  0.90948772  0.5883484
#> [12,]  1.4142136 -0.5084263 -1.2152872 -1.28869649  2.46450265  0.5883484
#> [13,]  1.4142136 -1.1364823 -1.2152872 -1.32353685  1.16560783  0.5883484
#> [14,] -0.7071068 -1.1364823 -1.2152872 -0.95241130  0.21430458 -0.7844645
#> [15,] -0.7071068 -0.5084263  0.9114654  0.44044560 -0.79188155 -0.7844645
#> [16,] -0.7071068  0.7476858  0.9114654  0.78884917  0.26918746 -0.7844645
#> [17,] -0.7071068 -0.5084263  0.9114654  1.16754871 -0.91994160 -0.7844645
#> [18,]  1.4142136 -0.5084263 -1.2152872 -0.95089651  1.54978798  1.9611614
#> [19,]  1.4142136 -0.5084263 -1.2152872 -1.14176107  0.34236463  1.9611614
#> [20,]  1.4142136  3.2599099  0.9114654  0.41772362 -0.07840411  1.9611614
#> [21,]  1.4142136 -0.5084263 -1.2152872 -0.94559471  0.96437060  0.5883484
#>                hp         qsec         vs          wt
#>  [1,] -0.63370603  0.822163634  1.1547005 -0.11549090
#>  [2,]  1.25541762 -1.217061127 -0.8660254  0.20374701
#>  [3,] -0.84360866  2.782085210  1.1547005 -0.17394291
#>  [4,] -0.45179042  0.516279920  1.1547005  0.08684299
#>  [5,]  0.34583957 -0.333397064 -0.8660254  0.65337787
#>  [6,]  0.34583957 -0.220106800 -0.8660254  0.34762889
#>  [7,]  0.34583957  0.006473729 -0.8660254  0.39259197
#>  [8,]  0.69567728 -0.004855297 -0.8660254  1.71450670
#>  [9,]  0.83561237 -0.095487509 -0.8660254  1.87097824
#> [10,]  1.04551499 -0.322068038 -0.8660254  1.79993656
#> [11,] -1.24942040  0.839157173  1.1547005 -1.02824154
#> [12,] -1.44532952  0.301028417  1.1547005 -1.55430964
#> [13,] -1.26341391  1.082731242  1.1547005 -1.35647207
#> [14,] -0.81562164  1.145040888  1.1547005 -0.78993719
#> [15,] -0.07396569 -0.390042196 -0.8660254  0.08234668
#> [16,]  1.25541762 -1.460635196 -0.8660254  0.44654768
#> [17,]  0.27587203 -0.531655027 -0.8660254  0.45104398
#> [18,] -0.89958269 -0.729912990 -0.8660254 -1.08219724
#> [19,] -0.59172550 -0.616622725  1.1547005 -1.64603434
#> [20,]  2.51483339 -1.919460767 -0.8660254  0.20374701
#> [21,] -0.64769954  0.346344523  1.1547005 -0.50666974
#> attr(,"center")
#>  [1]   0.3333333   2.8095238   6.2857143 245.8476190   3.5828571   3.5714286
#>  [7] 155.2857143  17.9885714   0.4285714   3.3434286
#> attr(,"scale")
#>  [1]   0.4714045   1.5922146   1.8808017 132.0307923   0.5466185   0.7284314
#>  [7]  71.4617066   1.7653768   0.4948717   1.1120233
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 45
#> 
#> $lambda.min
#> [1] 0.285523
#> 
#> $null.dev
#> [1] 47.61347
#> 
#> $Bias
#> [1] 2.43818
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 11.22667