Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 36.538359 33.180594 28.636005 24.838693 21.772869 19.327291 17.406282
#>   [8] 15.927630 14.820769 14.025196 13.489100 13.168182 13.024618 13.026171
#>  [15] 13.145420 13.205752 13.035556 12.898731 12.898731 12.898731 12.882792
#>  [22] 12.657289 12.131214 11.408914  9.991396  9.693777  9.661844  9.025864
#>  [29]  8.080839  7.983146  7.722165  7.602398  7.484219  6.436239  5.815109
#>  [36]  5.838199  5.915068  5.998282  6.037744  6.460591  6.110045  7.340732
#>  [43]  7.977790  7.831837  7.646192  7.671839  7.678235  7.426243  7.681202
#>  [50]  7.723310  7.723495  7.714998  7.841663  7.841515  7.841413  8.346555
#>  [57]  8.328605  9.025546  9.517521  9.598584 10.107055 11.466908 11.466915
#>  [64] 11.788074 11.694982 11.695188 11.670715 11.613198 11.613286 11.950475
#>  [71] 12.612098 12.611967 12.286176 12.310397 12.645822 12.648668 12.649196
#>  [78] 12.649700 12.650181 13.832571 13.833895 13.850015 13.851078 13.851975
#>  [85] 13.852930 13.895304 13.896401 13.897455 13.898501 13.908132 13.909042
#>  [92] 13.909947 13.914713 13.981202 13.982150 13.953572 13.954084 13.959502
#>  [99] 13.960545 13.961544
#> 
#> $cvse
#>   [1] 9.922191 9.646592 8.606344 7.628386 6.753652 5.970933 5.273344 4.658038
#>   [9] 4.126006 3.681719 3.332159 3.084579 2.942807 2.903432 2.954529 2.984623
#>  [17] 2.953391 2.919996 2.919996 2.919996 2.920868 2.887529 2.848729 2.795134
#>  [25] 2.836863 2.821006 2.813564 2.812498 2.655338 2.650287 2.651004 2.647666
#>  [33] 2.631468 2.492218 2.048381 2.009484 2.002041 1.998491 1.996228 1.754803
#>  [41] 1.770825 1.929376 2.224786 2.214836 2.218662 2.219109 2.221676 2.125180
#>  [49] 2.148481 2.154587 2.154626 2.156112 2.381477 2.381402 2.381341 2.431665
#>  [57] 2.543026 2.927526 2.887248 2.893889 2.844936 3.624077 3.623604 3.627117
#>  [65] 3.634392 3.634050 3.637405 3.663173 3.662899 3.700573 3.670920 3.670691
#>  [73] 3.677438 3.674170 3.912839 3.926273 3.926777 3.927265 3.927737 4.130227
#>  [81] 4.130822 4.129447 4.129877 4.130372 4.130858 4.134915 4.135407 4.135887
#>  [89] 4.136351 4.135866 4.136312 4.136735 4.135835 4.144689 4.145085 4.140932
#>  [97] 4.141378 4.141546 4.141898 4.142240
#> 
#> $fold
#>  [1]  1  4  9  3  2  4 10  3  8  2  1  9  6  7  1  7 10  6  5  5  8
#> 
#> $lambda
#>   [1] 5.171986257 4.823411693 4.498329888 4.195157509 3.912417934 3.648734060
#>   [7] 3.402821596 3.173482809 2.959600689 2.760133508 2.574109747 2.400623366
#>  [13] 2.238829385 2.087939778 1.947219625 1.815983539 1.693592327 1.579449872
#>  [19] 1.473000237 1.373724951 1.281140487 1.194795906 1.114270661 1.039172548
#>  [25] 0.969135796 0.903819287 0.842904892 0.786095923 0.733115689 0.683706145
#>  [31] 0.637626639 0.594652738 0.554575134 0.517198626 0.482341169 0.449832989
#>  [37] 0.419515751 0.391241793 0.364873406 0.340282160 0.317348281 0.295960069
#>  [43] 0.276013351 0.257410975 0.240062337 0.223882939 0.208793978 0.194721963
#>  [49] 0.181598354 0.169359233 0.157944987 0.147300024 0.137372495 0.128114049
#>  [55] 0.119479591 0.111427066 0.103917255 0.096913580 0.090381929 0.084290489
#>  [61] 0.078609592 0.073311569 0.068370615 0.063762664 0.059465274 0.055457513
#>  [67] 0.051719863 0.048234117 0.044983299 0.041951575 0.039124179 0.036487341
#>  [73] 0.034028216 0.031734828 0.029596007 0.027601335 0.025741097 0.024006234
#>  [79] 0.022388294 0.020879398 0.019472196 0.018159835 0.016935923 0.015794499
#>  [85] 0.014730002 0.013737250 0.012811405 0.011947959 0.011142707 0.010391725
#>  [91] 0.009691358 0.009038193 0.008429049 0.007860959 0.007331157 0.006837061
#>  [97] 0.006376266 0.005946527 0.005545751 0.005171986
#> 
#> $fit
#> $beta
#>               5.1720     4.8234     4.4983     4.1952    3.9124    3.6487
#> (Intercept) 18.81905 20.6957537 22.5155543 24.2127064 25.795476 27.271572
#> am           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> cyl          0.00000 -0.2897855 -0.5707841 -0.8328444 -1.077243 -1.305169
#> disp         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> wt           0.00000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#>                3.4028    3.1735    2.9596   2.7601    2.5741    2.4006
#> (Intercept) 28.648185 29.932018 31.129325 32.24594 33.287295 34.258468
#> am           0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> cyl         -1.517734 -1.715973 -1.900852 -2.07327 -2.234068 -2.384028
#> disp         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#>                2.2388   2.0879    1.9472    1.8160    1.6936    1.5794
#> (Intercept) 35.164187 36.00886 36.796612 37.531268 38.071084 38.071084
#> am           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.523882 -2.65431 -2.775947 -2.889387 -2.972741 -2.972741
#> disp         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#>                1.4730    1.3737    1.2811     1.1948    1.1143    1.0392
#> (Intercept) 38.071084 38.071084 38.071084 37.8130987 38.394423 38.654437
#> am           0.000000  0.000000  0.000000  0.2379201  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> cyl         -2.972741 -2.972741 -2.972741 -2.9451508 -2.431759 -1.994662
#> disp         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.0000000 -1.149745 -2.078331
#>                0.9691    0.9038    0.8429    0.7861    0.7331    0.6837
#> (Intercept) 38.763426 38.847122 38.847424 38.847603 38.847708 38.847771
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.629812 -1.666199 -1.666331 -1.666408 -1.666454 -1.666481
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.820972 -2.775319 -2.775154 -2.775057 -2.774999 -2.774965
#>                0.6376    0.5947    0.5546    0.5172    0.4823    0.4498
#> (Intercept) 38.847807 38.847829 38.847842 38.847849 38.847854 38.847857
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.666497 -1.666506 -1.666512 -1.666515 -1.666517 -1.666518
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.774945 -2.774934 -2.774927 -2.774922 -2.774920 -2.774919
#>                0.4195    0.3912    0.3649    0.3403      0.3173     0.2960
#> (Intercept) 38.847858 38.847859 38.847860 38.847860 38.73379623 38.5168318
#> am           0.000000  0.000000  0.000000  0.000000  0.09584128  0.2798369
#> carb         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> cyl         -1.666519 -1.666519 -1.666520 -1.666520 -1.67019394 -1.6781232
#> disp         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> wt          -2.774918 -2.774917 -2.774917 -2.774917 -2.74309676 -2.6809101
#>                 0.2760     0.2574     0.2401     0.2239     0.2088    0.1947
#> (Intercept) 38.3146942 35.3254483 35.3258783 35.3262420 35.3265393 35.326776
#> am           0.4510349  2.7311712  2.7311701  2.7311045  2.7310106  2.730910
#> carb         0.0000000 -0.6155044 -0.6154309 -0.6153664 -0.6153121 -0.615268
#> cyl         -1.6853862 -1.4839440 -1.4841505 -1.4842899 -1.4843819 -1.484441
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> wt          -2.6231927 -1.8082386 -1.8080301 -1.8079179 -1.8078662 -1.807851
#>                0.1816     0.1694     0.1579       0.1473     0.1374    0.1281
#> (Intercept) 35.326961 35.3271018 35.3272078 35.324336937 32.9841109 32.982216
#> am           2.730814  2.7307287  2.7306567  2.730471750  3.1767442  3.177634
#> carb        -0.615233 -0.6152059 -0.6151852 -0.614971139 -0.6376959 -0.637853
#> cyl         -1.484477 -1.4844990 -1.4845107 -1.484382997 -1.2377866 -1.237656
#> disp         0.000000  0.0000000  0.0000000  0.000000000  0.0000000  0.000000
#> drat         0.000000  0.0000000  0.0000000  0.000000000  0.0000000  0.000000
#> gear         0.000000  0.0000000  0.0000000  0.000000000  0.0000000  0.000000
#> hp           0.000000  0.0000000  0.0000000  0.000000000  0.0000000  0.000000
#> qsec         0.000000  0.0000000  0.0000000  0.000000000  0.0000000  0.000000
#> vs           0.000000  0.0000000  0.0000000  0.002232981  1.0167801  1.017244
#> wt          -1.807856 -1.8078719 -1.8078917 -1.807701543 -1.7253866 -1.725077
#>                 0.1195     0.1114     0.1039     0.0969    0.0904     0.0843
#> (Intercept) 32.9807064 32.9795050 32.9785493 32.9777891 32.977185 32.9767041
#> am           3.1783428  3.1789077  3.1793571  3.1797144  3.179998  3.1802238
#> carb        -0.6379773 -0.6380757 -0.6381535 -0.6382152 -0.638264 -0.6383028
#> cyl         -1.2375531 -1.2374725 -1.2374088 -1.2373584 -1.237318 -1.2372867
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> vs           1.0176152  1.0179119  1.0181487  1.0183376  1.018488  1.0186079
#> wt          -1.7248281 -1.7246289 -1.7244698 -1.7243430 -1.724242 -1.7241620
#>                 0.0786     0.0733     0.0684      0.0638      0.0595
#> (Intercept) 32.9763221 32.9760185 32.9757771 35.48877360 35.49032414
#> am           3.1804029  3.1805452  3.1806581  2.52403576  2.52365543
#> carb        -0.6383335 -0.6383579 -0.6383772  0.00000000  0.00000000
#> cyl         -1.2372615 -1.2372415 -1.2372255 -1.46915906 -1.46941256
#> disp         0.0000000  0.0000000  0.0000000  0.01714869  0.01714826
#> drat         0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> gear         0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> hp           0.0000000  0.0000000  0.0000000 -0.01798590 -0.01798327
#> qsec         0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> vs           1.0187033  1.0187792  1.0188396  1.34070777  1.34023600
#> wt          -1.7240983 -1.7240478 -1.7240077 -3.02541640 -3.02538977
#>                  0.0555      0.0517      0.0482      0.0450      0.0420
#> (Intercept) 35.49165646 35.49279320 35.49375523 35.49456175 35.49523039
#> am           2.52332096  2.52302817  2.52277314  2.52255221  2.52236197
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl         -1.46962873 -1.46981156 -1.46996473 -1.47009161 -1.47019527
#> disp         0.01714765  0.01714689  0.01714601  0.01714504  0.01714402
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.01798097 -0.01797895 -0.01797720 -0.01797569 -0.01797439
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           1.33981545  1.33944194  1.33911148  1.33882032  1.33856492
#> wt          -3.02535083 -3.02530203 -3.02524555 -3.02518325 -3.02511676
#>                  0.0391      0.0365      0.0340      0.0317      0.0296
#> (Intercept) 35.49577725 35.49621701 35.49656300 35.49682733 35.49702089
#> am           2.52219928  2.52206122  2.52194511  2.52184848  2.52176909
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl         -1.47027851 -1.47034390 -1.47039372 -1.47043008 -1.47045484
#> disp         0.01714295  0.01714186  0.01714076  0.01713966  0.01713858
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.01797328 -0.01797234 -0.01797156 -0.01797090 -0.01797037
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           1.33834195  1.33814833  1.33798115  1.33783773  1.33771560
#> wt          -3.02504747 -3.02497658 -3.02490510 -3.02483387 -3.02476360
#>                  0.0276      0.0257      0.0240      0.0224      0.0209
#> (Intercept) 35.49715352 35.49723404 35.49727033 35.49418496 34.90164178
#> am           2.52170487  2.52165396  2.52161465  2.52120250  2.46997789
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl         -1.47046968 -1.47047612 -1.47047550 -1.47041273 -1.44325006
#> disp         0.01713753  0.01713650  0.01713552  0.01713366  0.01722124
#> drat         0.00000000  0.00000000  0.00000000  0.00072400  0.12156122
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.01796994 -0.01796961 -0.01796935 -0.01796825 -0.01823713
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           1.33761245  1.33752620  1.33745492  1.33731940  1.36646564
#> wt          -3.02469485 -3.02462809 -3.02456369 -3.02438199 -3.02007248
#>                  0.0195      0.0182      0.0169        0.0158      0.0147
#> (Intercept) 34.90282132 34.90405498 34.90532508 34.9000584313 33.39466056
#> am           2.46997914  2.46996697  2.46994327  2.4701658094  2.57335024
#> carb         0.00000000  0.00000000  0.00000000  0.0000000000  0.00000000
#> cyl         -1.44341069 -1.44357715 -1.44374725 -1.4439384057 -1.40721030
#> disp         0.01722541  0.01722930  0.01723291  0.0172391510  0.01766674
#> drat         0.12146231  0.12135949  0.12125415  0.1211574508  0.13689022
#> gear         0.00000000  0.00000000  0.00000000  0.0000000000  0.00000000
#> hp          -0.01823669 -0.01823611 -0.01823542 -0.0182325454 -0.01763400
#> qsec         0.00000000  0.00000000  0.00000000  0.0004034629  0.07445510
#> vs           1.36650404  1.36651404  1.36649942  1.3657136720  1.27623332
#> wt          -3.02035855 -3.02062546 -3.02087361 -3.0215178427 -3.11333289
#>                  0.0137      0.0128      0.0119      0.0111      0.0104
#> (Intercept) 33.39459334 33.20412513 33.20328139 33.20256806 33.20197682
#> am           2.57338330  2.45890970  2.45906276  2.45921595  2.45936834
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl         -1.40745747 -1.39720225 -1.39730424 -1.39741148 -1.39752297
#> disp         0.01767080  0.01787128  0.01787550  0.01787949  0.01788324
#> drat         0.13671376  0.12423675  0.12416707  0.12409636  0.12402502
#> gear         0.00000000  0.09911034  0.09907328  0.09902677  0.09897203
#> hp          -0.01763041 -0.01823220 -0.01823061 -0.01822893 -0.01822716
#> qsec         0.07457556  0.06930825  0.06941251  0.06951234  0.06960767
#> vs           1.27586352  1.27907494  1.27896506  1.27884674  1.27872138
#> wt          -3.11373311 -3.11851408 -3.11890896 -3.11928336 -3.11963746
#>                  0.0097      0.0090      0.0084      0.0079        0.0073
#> (Intercept) 33.20149945 33.20112790 33.20085430 33.20067101 33.2029689047
#> am           2.45951916  2.45966771  2.45981338  2.45995566  2.4600941110
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.0002513809
#> cyl         -1.39763781 -1.39775516 -1.39787424 -1.39799434 -1.3982457938
#> disp         0.01788676  0.01789005  0.01789314  0.01789601  0.0178987876
#> drat         0.12395345  0.12388197  0.12381089  0.12374051  0.1234508262
#> gear         0.09891016  0.09884221  0.09876912  0.09869179  0.0984677550
#> hp          -0.01822532 -0.01822344 -0.01822152 -0.01821958 -0.0182204794
#> qsec         0.06969843  0.06978460  0.06986619  0.06994323  0.0699913272
#> vs           1.27859025  1.27845457  1.27831541  1.27817380  1.2780374811
#> wt          -3.11997152 -3.12028588 -3.12058093 -3.12085712 -3.1211525365
#>                  0.0068      0.0064      0.0059      0.0055      0.0052
#> (Intercept) 33.33223313 33.33256125 33.33288672 33.33320901 33.33352763
#> am           2.42108901  2.42104897  2.42101037  2.42097314  2.42093723
#> carb         0.04559776  0.04573894  0.04587551  0.04600759  0.04613533
#> cyl         -1.41195130 -1.41208782 -1.41222068 -1.41234990 -1.41247554
#> disp         0.01864243  0.01864678  0.01865099  0.01865506  0.01865899
#> drat         0.10497463  0.10486963  0.10476775  0.10466892  0.10457308
#> gear         0.09326033  0.09318880  0.09311892  0.09305070  0.09298415
#> hp          -0.01904545 -0.01904658 -0.01904766 -0.01904869 -0.01904968
#> qsec         0.07683230  0.07690859  0.07698235  0.07705365  0.07712256
#> vs           1.27520868  1.27507676  1.27494808  1.27482265  1.27470044
#> wt          -3.19789323 -3.19833079 -3.19875384 -3.19916281 -3.19955810
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   6  60  45  39   9   1   1   1   1   1   1   1   1   1   1
#>  [37]   1   1   1   1  12  14  13  43   1   1   1   1   1   1   1   2  38   1
#>  [55]   1   1   1   1   1   1   1   1   1 135   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   2  55   1   1   1   2  47   1  54   1   1   1
#>  [91]   1   1   1   1   1  77   1   1   1   1
#> 
#> $lambda
#>   [1] 5.171986257 4.823411693 4.498329888 4.195157509 3.912417934 3.648734060
#>   [7] 3.402821596 3.173482809 2.959600689 2.760133508 2.574109747 2.400623366
#>  [13] 2.238829385 2.087939778 1.947219625 1.815983539 1.693592327 1.579449872
#>  [19] 1.473000237 1.373724951 1.281140487 1.194795906 1.114270661 1.039172548
#>  [25] 0.969135796 0.903819287 0.842904892 0.786095923 0.733115689 0.683706145
#>  [31] 0.637626639 0.594652738 0.554575134 0.517198626 0.482341169 0.449832989
#>  [37] 0.419515751 0.391241793 0.364873406 0.340282160 0.317348281 0.295960069
#>  [43] 0.276013351 0.257410975 0.240062337 0.223882939 0.208793978 0.194721963
#>  [49] 0.181598354 0.169359233 0.157944987 0.147300024 0.137372495 0.128114049
#>  [55] 0.119479591 0.111427066 0.103917255 0.096913580 0.090381929 0.084290489
#>  [61] 0.078609592 0.073311569 0.068370615 0.063762664 0.059465274 0.055457513
#>  [67] 0.051719863 0.048234117 0.044983299 0.041951575 0.039124179 0.036487341
#>  [73] 0.034028216 0.031734828 0.029596007 0.027601335 0.025741097 0.024006234
#>  [79] 0.022388294 0.020879398 0.019472196 0.018159835 0.016935923 0.015794499
#>  [85] 0.014730002 0.013737250 0.012811405 0.011947959 0.011142707 0.010391725
#>  [91] 0.009691358 0.009038193 0.008429049 0.007860959 0.007331157 0.006837061
#>  [97] 0.006376266 0.005946527 0.005545751 0.005171986
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 22
#> 
#> $loss
#>   [1] 698.27238 594.62901 504.27105 429.00259 366.63443 315.27759 273.30323
#>   [8] 239.30773 212.08258 190.58806 173.93052 161.34263 152.16626 145.83763
#>  [15] 141.87446 139.86476 139.42539 139.42539 139.42539 139.42539 139.42539
#>  [22] 133.75691  90.28534  69.34365  64.66527  64.63038  64.63038  64.63038
#>  [29]  64.63038  64.63038  64.63038  64.63038  64.63038  64.63038  64.63038
#>  [36]  64.63038  64.63038  64.63038  64.63038  64.63038  64.03157  63.02129
#>  [43]  62.24582  52.43430  52.43430  52.43429  52.43429  52.43429  52.43429
#>  [50]  52.43429  52.43429  52.42758  50.89973  50.89972  50.89972  50.89971
#>  [57]  50.89971  50.89971  50.89971  50.89971  50.89971  50.89971  50.89971
#>  [64]  46.84433  46.84432  46.84431  46.84431  46.84431  46.84430  46.84430
#>  [71]  46.84430  46.84430  46.84430  46.84430  46.84430  46.84430  46.84430
#>  [78]  46.84430  46.84391  46.81210  46.81210  46.81209  46.81208  46.81154
#>  [85]  46.76325  46.76323  46.74225  46.74224  46.74223  46.74222  46.74221
#>  [92]  46.74220  46.74220  46.74219  46.74205  46.72932  46.72931  46.72930
#>  [99]  46.72929  46.72927
#> 
#> $linear.predictors
#>      5.1720   4.8234   4.4983   4.1952   3.9124   3.6487   3.4028   3.1735
#> 1  18.81905 18.95704 19.09085 19.21564 19.33202 19.44056 19.54178 19.63618
#> 2  18.81905 18.95704 19.09085 19.21564 19.33202 19.44056 19.54178 19.63618
#> 3  18.81905 18.37747 17.94928 17.54995 17.17754 16.83022 16.50631 16.20423
#> 4  18.81905 18.95704 19.09085 19.21564 19.33202 19.44056 19.54178 19.63618
#> 5  18.81905 18.37747 17.94928 17.54995 17.17754 16.83022 16.50631 16.20423
#> 6  18.81905 19.53661 20.23242 20.88133 21.48651 22.05090 22.57725 23.06812
#> 7  18.81905 18.95704 19.09085 19.21564 19.33202 19.44056 19.54178 19.63618
#> 8  18.81905 18.37747 17.94928 17.54995 17.17754 16.83022 16.50631 16.20423
#> 9  18.81905 18.37747 17.94928 17.54995 17.17754 16.83022 16.50631 16.20423
#> 10 18.81905 18.37747 17.94928 17.54995 17.17754 16.83022 16.50631 16.20423
#> 11 18.81905 18.37747 17.94928 17.54995 17.17754 16.83022 16.50631 16.20423
#> 12 18.81905 19.53661 20.23242 20.88133 21.48651 22.05090 22.57725 23.06812
#> 13 18.81905 19.53661 20.23242 20.88133 21.48651 22.05090 22.57725 23.06812
#> 14 18.81905 18.37747 17.94928 17.54995 17.17754 16.83022 16.50631 16.20423
#> 15 18.81905 18.37747 17.94928 17.54995 17.17754 16.83022 16.50631 16.20423
#> 16 18.81905 18.37747 17.94928 17.54995 17.17754 16.83022 16.50631 16.20423
#> 17 18.81905 19.53661 20.23242 20.88133 21.48651 22.05090 22.57725 23.06812
#> 18 18.81905 19.53661 20.23242 20.88133 21.48651 22.05090 22.57725 23.06812
#> 19 18.81905 19.53661 20.23242 20.88133 21.48651 22.05090 22.57725 23.06812
#> 20 18.81905 18.37747 17.94928 17.54995 17.17754 16.83022 16.50631 16.20423
#> 21 18.81905 18.37747 17.94928 17.54995 17.17754 16.83022 16.50631 16.20423
#>      2.9596   2.7601   2.5741   2.4006   2.2388   2.0879   1.9472   1.8160
#> 1  19.72422 19.80632 19.88289 19.95430 20.02090 20.08300 20.14093 20.19495
#> 2  19.72422 19.80632 19.88289 19.95430 20.02090 20.08300 20.14093 20.19495
#> 3  15.92251 15.65978 15.41475 15.18624 14.97313 14.77439 14.58903 14.41617
#> 4  19.72422 19.80632 19.88289 19.95430 20.02090 20.08300 20.14093 20.19495
#> 5  15.92251 15.65978 15.41475 15.18624 14.97313 14.77439 14.58903 14.41617
#> 6  23.52592 23.95286 24.35102 24.72236 25.06866 25.39162 25.69282 25.97372
#> 7  19.72422 19.80632 19.88289 19.95430 20.02090 20.08300 20.14093 20.19495
#> 8  15.92251 15.65978 15.41475 15.18624 14.97313 14.77439 14.58903 14.41617
#> 9  15.92251 15.65978 15.41475 15.18624 14.97313 14.77439 14.58903 14.41617
#> 10 15.92251 15.65978 15.41475 15.18624 14.97313 14.77439 14.58903 14.41617
#> 11 15.92251 15.65978 15.41475 15.18624 14.97313 14.77439 14.58903 14.41617
#> 12 23.52592 23.95286 24.35102 24.72236 25.06866 25.39162 25.69282 25.97372
#> 13 23.52592 23.95286 24.35102 24.72236 25.06866 25.39162 25.69282 25.97372
#> 14 15.92251 15.65978 15.41475 15.18624 14.97313 14.77439 14.58903 14.41617
#> 15 15.92251 15.65978 15.41475 15.18624 14.97313 14.77439 14.58903 14.41617
#> 16 15.92251 15.65978 15.41475 15.18624 14.97313 14.77439 14.58903 14.41617
#> 17 23.52592 23.95286 24.35102 24.72236 25.06866 25.39162 25.69282 25.97372
#> 18 23.52592 23.95286 24.35102 24.72236 25.06866 25.39162 25.69282 25.97372
#> 19 23.52592 23.95286 24.35102 24.72236 25.06866 25.39162 25.69282 25.97372
#> 20 15.92251 15.65978 15.41475 15.18624 14.97313 14.77439 14.58903 14.41617
#> 21 15.92251 15.65978 15.41475 15.18624 14.97313 14.77439 14.58903 14.41617
#>      1.6936   1.5794   1.4730   1.3737   1.2811   1.1948   1.1143   1.0392
#> 1  20.23464 20.23464 20.23464 20.23464 20.23464 20.38011 20.79154 21.24124
#> 2  20.23464 20.23464 20.23464 20.23464 20.23464 20.14219 20.10744 20.00463
#> 3  14.28916 14.28916 14.28916 14.28916 14.28916 14.25189 14.98523 15.54768
#> 4  20.23464 20.23464 20.23464 20.23464 20.23464 20.14219 19.82575 19.49544
#> 5  14.28916 14.28916 14.28916 14.28916 14.28916 14.25189 14.83577 15.27750
#> 6  26.18012 26.18012 26.18012 26.18012 26.18012 26.03250 25.04569 24.12904
#> 7  20.23464 20.23464 20.23464 20.23464 20.23464 20.14219 19.84875 19.53700
#> 8  14.28916 14.28916 14.28916 14.28916 14.28916 14.25189 14.59432 14.84105
#> 9  14.28916 14.28916 14.28916 14.28916 14.28916 14.25189 12.90419 11.78590
#> 10 14.28916 14.28916 14.28916 14.28916 14.28916 14.25189 12.70414 11.42427
#> 11 14.28916 14.28916 14.28916 14.28916 14.28916 14.25189 12.79497 11.58846
#> 12 26.18012 26.18012 26.18012 26.18012 26.18012 26.27042 26.81055 27.31928
#> 13 26.18012 26.18012 26.18012 26.18012 26.18012 26.03250 25.83327 25.55270
#> 14 14.28916 14.28916 14.28916 14.28916 14.28916 14.25189 14.89325 15.38141
#> 15 14.28916 14.28916 14.28916 14.28916 14.28916 14.25189 14.99098 15.55807
#> 16 14.28916 14.28916 14.28916 14.28916 14.28916 14.25189 14.52533 14.71635
#> 17 26.18012 26.18012 26.18012 26.18012 26.18012 26.27042 26.44263 26.65422
#> 18 26.18012 26.18012 26.18012 26.18012 26.18012 26.27042 26.20693 26.22816
#> 19 26.18012 26.18012 26.18012 26.18012 26.18012 26.27042 26.92782 27.53127
#> 20 14.28916 14.28916 14.28916 14.28916 14.28916 14.48981 15.29566 16.10883
#> 21 14.28916 14.28916 14.28916 14.28916 14.28916 14.48981 14.83577 15.27750
#>      0.9691   0.9038   0.8429   0.7861   0.7331   0.6837   0.6376   0.5947
#> 1  21.59361 21.57859 21.57854 21.57850 21.57849 21.57847 21.57847 21.57846
#> 2  19.91513 19.92728 19.92732 19.92735 19.92736 19.92737 19.92738 19.92738
#> 3  16.02078 15.97043 15.97025 15.97014 15.97008 15.97004 15.97002 15.97001
#> 4  19.22399 19.24732 19.24741 19.24746 19.24749 19.24750 19.24751 19.24752
#> 5  15.65406 15.60964 15.60948 15.60938 15.60933 15.60930 15.60928 15.60926
#> 6  23.35811 23.44007 23.44037 23.44054 23.44064 23.44071 23.44074 23.44076
#> 7  19.28041 19.30283 19.30291 19.30296 19.30299 19.30300 19.30301 19.30302
#> 8  15.06165 15.02682 15.02670 15.02662 15.02658 15.02655 15.02654 15.02653
#> 9  10.91482 10.94710 10.94722 10.94729 10.94733 10.94735 10.94737 10.94738
#> 10 10.42398 10.46420 10.46434 10.46443 10.46448 10.46451 10.46453 10.46454
#> 11 10.64683 10.68345 10.68358 10.68366 10.68370 10.68373 10.68375 10.68376
#> 12 27.68831 27.70019 27.70023 27.70025 27.70027 27.70028 27.70028 27.70029
#> 13 25.29048 25.34116 25.34135 25.34146 25.34152 25.34156 25.34158 25.34159
#> 14 15.79511 15.74841 15.74824 15.74814 15.74808 15.74804 15.74802 15.74801
#> 15 16.03489 15.98431 15.98413 15.98402 15.98395 15.98392 15.98389 15.98388
#> 16 14.89240 14.86030 14.86019 14.86012 14.86008 14.86005 14.86004 14.86003
#> 17 26.78560 26.81208 26.81218 26.81224 26.81227 26.81229 26.81230 26.81231
#> 18 26.20730 26.24314 26.24327 26.24335 26.24339 26.24342 26.24344 26.24345
#> 19 27.97605 27.98327 27.98329 27.98331 27.98332 27.98332 27.98333 27.98333
#> 20 16.78245 16.71977 16.71954 16.71941 16.71933 16.71928 16.71925 16.71924
#> 21 15.65406 15.60964 15.60948 15.60938 15.60933 15.60930 15.60928 15.60926
#>      0.5546   0.5172   0.4823   0.4498   0.4195   0.3912   0.3649   0.3403
#> 1  21.57846 21.57846 21.57846 21.57846 21.57846 21.57846 21.57846 21.57846
#> 2  19.92738 19.92738 19.92738 19.92738 19.92738 19.92738 19.92738 19.92738
#> 3  15.97000 15.96999 15.96999 15.96999 15.96999 15.96999 15.96999 15.96999
#> 4  19.24752 19.24753 19.24753 19.24753 19.24753 19.24753 19.24753 19.24753
#> 5  15.60926 15.60925 15.60925 15.60925 15.60925 15.60925 15.60925 15.60925
#> 6  23.44077 23.44078 23.44079 23.44079 23.44079 23.44079 23.44079 23.44079
#> 7  19.30302 19.30302 19.30303 19.30303 19.30303 19.30303 19.30303 19.30303
#> 8  15.02652 15.02652 15.02652 15.02652 15.02652 15.02652 15.02652 15.02652
#> 9  10.94738 10.94738 10.94739 10.94739 10.94739 10.94739 10.94739 10.94739
#> 10 10.46454 10.46455 10.46455 10.46455 10.46455 10.46455 10.46455 10.46455
#> 11 10.68376 10.68377 10.68377 10.68377 10.68377 10.68377 10.68377 10.68377
#> 12 27.70029 27.70029 27.70029 27.70029 27.70029 27.70029 27.70029 27.70029
#> 13 25.34160 25.34160 25.34161 25.34161 25.34161 25.34161 25.34161 25.34161
#> 14 15.74800 15.74800 15.74800 15.74800 15.74800 15.74799 15.74799 15.74799
#> 15 15.98387 15.98387 15.98387 15.98386 15.98386 15.98386 15.98386 15.98386
#> 16 14.86003 14.86002 14.86002 14.86002 14.86002 14.86002 14.86002 14.86002
#> 17 26.81231 26.81231 26.81231 26.81232 26.81232 26.81232 26.81232 26.81232
#> 18 26.24345 26.24345 26.24346 26.24346 26.24346 26.24346 26.24346 26.24346
#> 19 27.98333 27.98333 27.98333 27.98333 27.98333 27.98333 27.98333 27.98333
#> 20 16.71923 16.71922 16.71922 16.71922 16.71922 16.71922 16.71922 16.71921
#> 21 15.60926 15.60925 15.60925 15.60925 15.60925 15.60925 15.60925 15.60925
#>      0.3173   0.2960   0.2760   0.2574   0.2401   0.2239   0.2088   0.1947
#> 1  21.62156 21.70395 21.78065 21.95335 21.95338 21.95340 21.95340 21.95340
#> 2  19.89358 19.82897 19.76881 19.99279 19.99273 19.99268 19.99265 19.99262
#> 3  15.93599 15.86952 15.80782 16.00255 16.00219 16.00195 16.00180 16.00171
#> 4  19.22152 19.17214 19.12613 19.54977 19.54976 19.54974 19.54972 19.54970
#> 5  15.57939 15.52100 15.46681 14.53647 14.53628 14.53619 14.53615 14.53615
#> 6  23.41227 23.35947 23.31009 22.46271 22.46312 22.46341 22.46361 22.46375
#> 7  19.27638 19.22576 19.17859 17.73943 17.73963 17.73980 17.73994 17.74005
#> 8  15.00334 14.95801 14.91594 14.77224 14.77203 14.77189 14.77181 14.77177
#> 9  10.97099 11.01707 11.05984 11.49863 11.49879 11.49889 11.49894 11.49896
#> 10 10.49369 10.55059 10.60341 11.18399 11.18420 11.18431 11.18437 11.18439
#> 11 10.71039 10.76238 10.81064 11.32684 11.32703 11.32714 11.32719 11.32721
#> 12 27.71876 27.75451 27.78773 27.96953 27.96962 27.96967 27.96969 27.96971
#> 13 25.29129 25.19590 25.10698 24.31686 24.31705 24.31720 24.31731 24.31739
#> 14 15.71654 15.65504 15.59797 15.85789 15.85755 15.85732 15.85717 15.85708
#> 15 15.94971 15.88292 15.82094 16.01159 16.01123 16.01099 16.01084 16.01075
#> 16 14.83875 14.79715 14.75854 14.04824 14.04812 14.04805 14.04803 14.04803
#> 17 26.84097 26.89661 26.94831 28.00640 28.00648 28.00650 28.00649 28.00646
#> 18 26.27863 26.34703 26.41055 27.02020 27.02040 27.02051 27.02056 27.02059
#> 19 27.99856 28.02796 28.05529 28.15397 28.15403 28.15407 28.15410 28.15411
#> 20 16.77247 16.87320 16.96712 17.99093 17.99067 17.99046 17.99031 17.99020
#> 21 15.67523 15.80083 15.91784 14.80562 14.80573 14.80583 14.80592 14.80599
#>      0.1816   0.1694   0.1579   0.1473   0.1374   0.1281   0.1195   0.1114
#> 1  21.95339 21.95339 21.95338 21.95245 21.66284 21.66280 21.66277 21.66275
#> 2  19.99261 19.99259 19.99259 19.99354 20.38936 20.38955 20.38970 20.38982
#> 3  16.00165 16.00162 16.00160 16.00084 15.87110 15.87100 15.87092 15.87085
#> 4  19.54968 19.54967 19.54965 19.55065 19.96664 19.96691 19.96712 19.96729
#> 5  14.53616 14.53618 14.53621 14.53589 14.37140 14.37103 14.37074 14.37050
#> 6  22.46384 22.46390 22.46394 22.46484 22.33938 22.33914 22.33895 22.33879
#> 7  17.74014 17.74021 17.74026 17.74189 18.08806 18.08785 18.08768 18.08756
#> 8  14.77175 14.77174 14.77174 14.77125 14.64677 14.64662 14.64650 14.64640
#> 9  11.49896 11.49896 11.49895 11.49896 11.47275 11.47290 11.47302 11.47312
#> 10 11.18440 11.18439 11.18438 11.18442 11.17254 11.17274 11.17290 11.17304
#> 11 11.32722 11.32721 11.32720 11.32722 11.30884 11.30902 11.30917 11.30928
#> 12 27.96971 27.96971 27.96971 27.97013 28.16460 28.16477 28.16490 28.16501
#> 13 24.31745 24.31750 24.31753 24.31808 24.15897 24.15867 24.15843 24.15824
#> 14 15.85702 15.85699 15.85697 15.85622 15.73307 15.73299 15.73293 15.73288
#> 15 16.01069 16.01066 16.01064 16.00988 15.87972 15.87962 15.87954 15.87947
#> 16 14.04804 14.04806 14.04808 14.04781 13.90555 13.90526 13.90503 13.90485
#> 17 28.00643 28.00640 28.00637 28.00664 28.25017 28.25059 28.25093 28.25120
#> 18 27.02059 27.02058 27.02056 27.01885 26.24199 26.24186 26.24175 26.24167
#> 19 28.15411 28.15411 28.15411 28.15451 28.34059 28.34072 28.34083 28.34092
#> 20 17.99012 17.99006 17.99002 17.98945 18.23830 18.23870 18.23901 18.23926
#> 21 14.80604 14.80609 14.80612 14.80648 14.99736 14.99725 14.99717 14.99710
#>      0.1039   0.0969   0.0904   0.0843   0.0786   0.0733   0.0684   0.0638
#> 1  21.66273 21.66271 21.66270 21.66269 21.66268 21.66268 21.66267 22.03661
#> 2  20.38992 20.39000 20.39006 20.39011 20.39015 20.39018 20.39020 20.73373
#> 3  15.87080 15.87075 15.87072 15.87069 15.87066 15.87065 15.87063 16.35406
#> 4  19.96743 19.96753 19.96762 19.96769 19.96774 19.96779 19.96782 19.51652
#> 5  14.37031 14.37016 14.37004 14.36994 14.36986 14.36980 14.36976 14.70175
#> 6  22.33868 22.33858 22.33851 22.33845 22.33840 22.33837 22.33834 22.12866
#> 7  18.08746 18.08738 18.08731 18.08726 18.08722 18.08719 18.08717 18.26895
#> 8  14.64632 14.64626 14.64621 14.64617 14.64614 14.64611 14.64609 13.79157
#> 9  11.47320 11.47326 11.47331 11.47335 11.47338 11.47340 11.47342 12.25914
#> 10 11.17314 11.17322 11.17329 11.17334 11.17339 11.17342 11.17345 11.34707
#> 11 11.30937 11.30945 11.30951 11.30955 11.30959 11.30962 11.30964 10.97332
#> 12 28.16509 28.16516 28.16522 28.16526 28.16530 28.16532 28.16535 28.95372
#> 13 24.15809 24.15797 24.15788 24.15780 24.15774 24.15770 24.15766 23.81012
#> 14 15.73284 15.73280 15.73278 15.73275 15.73274 15.73272 15.73271 15.84143
#> 15 15.87942 15.87937 15.87934 15.87931 15.87929 15.87927 15.87925 15.85851
#> 16 13.90470 13.90458 13.90449 13.90442 13.90436 13.90431 13.90427 13.71340
#> 17 28.25142 28.25159 28.25172 28.25183 28.25192 28.25199 28.25204 27.79038
#> 18 26.24160 26.24155 26.24150 26.24147 26.24144 26.24142 26.24140 26.08805
#> 19 28.34099 28.34105 28.34109 28.34113 28.34115 28.34118 28.34119 28.49786
#> 20 18.23945 18.23961 18.23973 18.23983 18.23991 18.23997 18.24002 17.93988
#> 21 14.99705 14.99701 14.99698 14.99695 14.99693 14.99692 14.99691 14.59528
#>      0.0595   0.0555   0.0517   0.0482   0.0450   0.0420   0.0391   0.0365
#> 1  22.03654 22.03650 22.03648 22.03647 22.03647 22.03647 22.03649 22.03652
#> 2  20.73355 20.73338 20.73323 20.73309 20.73296 20.73285 20.73274 20.73264
#> 3  16.35398 16.35390 16.35382 16.35375 16.35367 16.35359 16.35352 16.35346
#> 4  19.51635 19.51620 19.51608 19.51597 19.51589 19.51581 19.51575 19.51570
#> 5  14.70185 14.70194 14.70201 14.70206 14.70210 14.70212 14.70214 14.70214
#> 6  22.12900 22.12930 22.12957 22.12981 22.13002 22.13021 22.13037 22.13051
#> 7  18.26885 18.26878 18.26873 18.26871 18.26870 18.26871 18.26873 18.26876
#> 8  13.79155 13.79155 13.79156 13.79158 13.79162 13.79166 13.79170 13.79176
#> 9  12.25914 12.25913 12.25911 12.25909 12.25906 12.25903 12.25900 12.25897
#> 10 11.34711 11.34713 11.34715 11.34717 11.34718 11.34719 11.34719 11.34719
#> 11 10.97340 10.97347 10.97353 10.97359 10.97364 10.97368 10.97371 10.97374
#> 12 28.95355 28.95340 28.95327 28.95315 28.95304 28.95295 28.95286 28.95279
#> 13 23.81045 23.81075 23.81100 23.81122 23.81142 23.81158 23.81172 23.81184
#> 14 15.84131 15.84120 15.84110 15.84102 15.84095 15.84090 15.84085 15.84081
#> 15 15.85839 15.85829 15.85820 15.85813 15.85806 15.85801 15.85798 15.85794
#> 16 13.71352 13.71362 13.71371 13.71378 13.71385 13.71390 13.71395 13.71398
#> 17 27.79025 27.79014 27.79005 27.78997 27.78990 27.78984 27.78979 27.78975
#> 18 26.08845 26.08881 26.08911 26.08938 26.08962 26.08982 26.08999 26.09014
#> 19 28.49784 28.49781 28.49778 28.49775 28.49771 28.49766 28.49762 28.49758
#> 20 17.93965 17.93943 17.93923 17.93905 17.93888 17.93872 17.93858 17.93844
#> 21 14.59527 14.59526 14.59526 14.59527 14.59528 14.59529 14.59531 14.59533
#>      0.0340   0.0317   0.0296   0.0276   0.0257   0.0240   0.0224   0.0209
#> 1  22.03654 22.03658 22.03661 22.03665 22.03669 22.03673 22.03673 22.02293
#> 2  20.73256 20.73248 20.73241 20.73234 20.73229 20.73223 20.73185 20.71048
#> 3  16.35339 16.35333 16.35327 16.35322 16.35316 16.35312 16.35296 16.35766
#> 4  19.51567 19.51564 19.51562 19.51560 19.51560 19.51559 19.51507 19.45455
#> 5  14.70214 14.70214 14.70213 14.70211 14.70209 14.70207 14.70206 14.69574
#> 6  22.13064 22.13075 22.13084 22.13092 22.13098 22.13104 22.13132 22.15062
#> 7  18.26880 18.26884 18.26889 18.26894 18.26900 18.26906 18.26950 18.33919
#> 8  13.79181 13.79187 13.79193 13.79199 13.79205 13.79211 13.79212 13.77990
#> 9  12.25893 12.25889 12.25886 12.25882 12.25879 12.25875 12.25860 12.24625
#> 10 11.34719 11.34719 11.34718 11.34718 11.34717 11.34716 11.34712 11.34024
#> 11 10.97377 10.97379 10.97381 10.97383 10.97384 10.97385 10.97401 10.98880
#> 12 28.95273 28.95267 28.95262 28.95258 28.95255 28.95252 28.95292 29.04228
#> 13 23.81194 23.81203 23.81209 23.81215 23.81219 23.81222 23.81226 23.79967
#> 14 15.84077 15.84075 15.84073 15.84071 15.84070 15.84069 15.84032 15.80128
#> 15 15.85792 15.85790 15.85789 15.85789 15.85788 15.85788 15.85781 15.86430
#> 16 13.71401 13.71404 13.71405 13.71407 13.71408 13.71408 13.71452 13.77132
#> 17 27.78972 27.78969 27.78967 27.78966 27.78964 27.78964 27.78949 27.77404
#> 18 26.09026 26.09036 26.09044 26.09051 26.09057 26.09061 26.09083 26.08632
#> 19 28.49753 28.49749 28.49745 28.49740 28.49736 28.49732 28.49689 28.43095
#> 20 17.93832 17.93821 17.93812 17.93803 17.93794 17.93787 17.93815 17.99503
#> 21 14.59534 14.59536 14.59538 14.59541 14.59543 14.59545 14.59547 14.54844
#>      0.0195   0.0182   0.0169   0.0158   0.0147   0.0137   0.0128   0.0119
#> 1  22.02273 22.02254 22.02236 22.02206 22.01416 22.01393 21.98496 21.98477
#> 2  20.71063 20.71077 20.71089 20.71109 20.70557 20.70561 20.70655 20.70668
#> 3  16.35784 16.35800 16.35814 16.35844 16.39962 16.39978 16.41015 16.41037
#> 4  19.45452 19.45449 19.45446 19.45463 19.46225 19.46218 19.45822 19.45821
#> 5  14.69591 14.69607 14.69622 14.69616 14.68086 14.68107 14.65416 14.65431
#> 6  22.15054 22.15047 22.15043 22.15159 22.28893 22.28923 22.32590 22.32600
#> 7  18.33883 18.33847 18.33813 18.33736 18.25354 18.25295 18.31834 18.31787
#> 8  13.77963 13.77939 13.77915 13.77913 13.82739 13.82723 13.81191 13.81176
#> 9  12.24641 12.24656 12.24671 12.24704 12.25550 12.25566 12.25946 12.25960
#> 10 11.34030 11.34036 11.34041 11.34051 11.32311 11.32315 11.31767 11.31769
#> 11 10.98878 10.98877 10.98876 10.98865 10.95292 10.95288 10.93397 10.93391
#> 12 29.04225 29.04220 29.04214 29.04212 29.06155 29.06137 29.01420 29.01419
#> 13 23.79972 23.79978 23.79986 23.80020 23.77530 23.77549 23.72894 23.72898
#> 14 15.80129 15.80129 15.80128 15.80118 15.78484 15.78476 15.80703 15.80702
#> 15 15.86423 15.86416 15.86409 15.86408 15.88755 15.88743 15.90016 15.90014
#> 16 13.77132 13.77131 13.77131 13.77079 13.70276 13.70268 13.66825 13.66817
#> 17 27.77402 27.77399 27.77397 27.77404 27.78865 27.78859 27.74074 27.74074
#> 18 26.08635 26.08640 26.08646 26.08660 26.04708 26.04724 26.09476 26.09471
#> 19 28.43116 28.43137 28.43157 28.43138 28.36676 28.36693 28.40954 28.40963
#> 20 17.99518 17.99531 17.99541 17.99519 18.04399 18.04408 18.08405 18.08422
#> 21 14.54837 14.54830 14.54825 14.54777 14.57766 14.57778 14.57104 14.57102
#>      0.0111   0.0104   0.0097   0.0090   0.0084   0.0079   0.0073   0.0068
#> 1  21.98458 21.98442 21.98426 21.98411 21.98397 21.98384 21.98411 22.02042
#> 2  20.70679 20.70690 20.70700 20.70708 20.70716 20.70723 20.70718 20.71157
#> 3  16.41058 16.41077 16.41095 16.41111 16.41126 16.41139 16.41125 16.42352
#> 4  19.45820 19.45818 19.45817 19.45815 19.45813 19.45811 19.45802 19.43446
#> 5  14.65447 14.65462 14.65477 14.65492 14.65506 14.65520 14.65538 14.68145
#> 6  22.32611 22.32621 22.32631 22.32641 22.32650 22.32659 22.32670 22.33701
#> 7  18.31740 18.31695 18.31650 18.31607 18.31565 18.31525 18.31512 18.33593
#> 8  13.81162 13.81148 13.81135 13.81123 13.81112 13.81101 13.81084 13.78381
#> 9  12.25974 12.25987 12.26000 12.26011 12.26023 12.26033 12.26054 12.29378
#> 10 11.31771 11.31773 11.31775 11.31778 11.31780 11.31783 11.31792 11.31824
#> 11 10.93384 10.93379 10.93374 10.93370 10.93367 10.93364 10.93360 10.90576
#> 12 29.01419 29.01417 29.01416 29.01414 29.01413 29.01410 29.01416 29.04169
#> 13 23.72903 23.72910 23.72918 23.72926 23.72935 23.72944 23.72959 23.71958
#> 14 15.80700 15.80697 15.80694 15.80691 15.80687 15.80683 15.80667 15.80838
#> 15 15.90011 15.90008 15.90004 15.89999 15.89995 15.89990 15.89963 15.89319
#> 16 13.66811 13.66805 13.66800 13.66795 13.66791 13.66788 13.66778 13.65314
#> 17 27.74075 27.74075 27.74077 27.74078 27.74080 27.74081 27.74078 27.70762
#> 18 26.09468 26.09465 26.09463 26.09461 26.09461 26.09460 26.09456 26.07721
#> 19 28.40972 28.40979 28.40987 28.40994 28.41000 28.41006 28.41018 28.41479
#> 20 18.08438 18.08451 18.08463 18.08474 18.08483 18.08490 18.08449 18.04164
#> 21 14.57100 14.57099 14.57099 14.57100 14.57101 14.57102 14.57150 14.59682
#>      0.0064   0.0059   0.0055   0.0052
#> 1  22.02044 22.02046 22.02048 22.02049
#> 2  20.71163 20.71169 20.71174 20.71179
#> 3  16.42366 16.42379 16.42392 16.42404
#> 4  19.43436 19.43427 19.43418 19.43409
#> 5  14.68163 14.68182 14.68199 14.68216
#> 6  22.33712 22.33722 22.33733 22.33743
#> 7  18.33570 18.33548 18.33526 18.33505
#> 8  13.78365 13.78350 13.78335 13.78321
#> 9  12.29396 12.29413 12.29430 12.29446
#> 10 11.31826 11.31828 11.31830 11.31831
#> 11 10.90565 10.90555 10.90545 10.90536
#> 12 29.04176 29.04182 29.04188 29.04194
#> 13 23.71963 23.71968 23.71972 23.71977
#> 14 15.80836 15.80833 15.80831 15.80829
#> 15 15.89313 15.89308 15.89302 15.89297
#> 16 13.65308 13.65302 13.65296 13.65291
#> 17 27.70752 27.70743 27.70734 27.70725
#> 18 26.07717 26.07713 26.07710 26.07706
#> 19 28.41484 28.41489 28.41494 28.41499
#> 20 18.04155 18.04146 18.04137 18.04129
#> 21 14.59690 14.59698 14.59706 14.59713
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.4 18.7 18.1 14.3 22.8 17.8 15.2 10.4 10.4 14.7 30.4 21.5 15.5 15.2
#> [16] 13.3 27.3 26.0 30.4 15.8 15.0
#> 
#> $X
#>               am        carb        cyl         disp        drat       gear
#>  [1,]  1.4142136  0.67183717 -0.2744106 -0.790131658  0.61944892  0.4850713
#>  [2,] -0.7071068 -1.16841248 -0.2744106 -0.005868635 -0.81814008 -0.7882408
#>  [3,] -0.7071068 -0.55499593  0.8781141  0.810405123 -0.69541907 -0.7882408
#>  [4,] -0.7071068 -1.16841248 -0.2744106 -0.269957204 -1.37915042 -0.7882408
#>  [5,] -0.7071068  0.67183717  0.8781141  0.810405123 -0.59022963 -0.7882408
#>  [6,] -0.7071068 -0.55499593 -1.4269354 -0.943783189  0.65451206  0.4850713
#>  [7,] -0.7071068  0.67183717 -0.2744106 -0.729311260  0.65451206  0.4850713
#>  [8,] -0.7071068  0.05842062  0.8781141  0.136579139 -0.83567165 -0.7882408
#>  [9,] -0.7071068  0.67183717  0.8781141  1.706705721 -1.08111367 -0.7882408
#> [10,] -0.7071068  0.67183717  0.8781141  1.610673514 -0.95839266 -0.7882408
#> [11,] -0.7071068  0.67183717  0.8781141  1.450619836 -0.55516648 -0.7882408
#> [12,]  1.4142136 -0.55499593 -1.4269354 -1.464757911  2.42520094  0.4850713
#> [13,] -0.7071068 -1.16841248 -1.4269354 -1.109438745  0.26881745 -0.7882408
#> [14,] -0.7071068 -0.55499593  0.8781141  0.474292399 -1.37915042 -0.7882408
#> [15,] -0.7071068 -0.55499593  0.8781141  0.362254825 -0.69541907 -0.7882408
#> [16,] -0.7071068  0.67183717  0.8781141  0.730378284  0.32141217 -0.7882408
#> [17,]  1.4142136 -1.16841248 -1.4269354 -1.438349054  0.93501723  0.4850713
#> [18,]  1.4142136 -0.55499593 -1.4269354 -1.107838209  1.54862229  1.7583833
#> [19,]  1.4142136 -0.55499593 -1.4269354 -1.309505843  0.39153847  1.7583833
#> [20,]  1.4142136  0.67183717  0.8781141  0.738380968  1.18045925  1.7583833
#> [21,]  1.4142136  3.12550337  0.8781141  0.338246773 -0.01168772  1.7583833
#>               hp        qsec         vs         wt
#>  [1,] -0.6828893 -0.64114811 -0.7844645 -0.6664922
#>  [2,] -0.6828893  0.89908409  1.2747549 -0.1067140
#>  [3,]  0.2082616 -0.35170850 -0.7844645  0.1049668
#>  [4,] -0.7514393  1.30223211  1.2747549  0.1237829
#>  [5,]  1.1679626 -0.96159910 -0.7844645  0.2272713
#>  [6,] -0.8885395  2.68740738  1.2747549 -0.1678663
#>  [7,] -0.5046591  0.61998161  1.2747549  0.1049668
#>  [8,]  0.2768117  0.15481081 -0.7844645  0.4248401
#>  [9,]  0.6195620  0.14447368 -0.7844645  1.8078215
#> [10,]  0.7566622  0.06177665 -0.7844645  1.9715213
#> [11,]  0.9623124 -0.14496593 -0.7844645  1.8971978
#> [12,] -1.4780701  0.42357616  1.2747549 -1.6119999
#> [13,] -0.8611194  1.19369226  1.2747549 -0.8123168
#> [14,] -0.1344887 -0.42923697 -0.7844645  0.1802311
#> [15,] -0.1344887 -0.20698870 -0.7844645  0.1002628
#> [16,]  1.1679626 -1.18384737 -0.7844645  0.4812883
#> [17,] -1.2861299  0.61998161  1.2747549 -1.3109427
#> [18,] -0.9433795 -0.51710256 -0.7844645 -1.1180780
#> [19,] -0.6417592 -0.41373128  1.2747549 -1.7079619
#> [20,]  1.4284529 -1.65418674 -0.7844645 -0.1490502
#> [21,]  2.4018638 -1.60250109 -0.7844645  0.2272713
#> attr(,"center")
#>  [1]   0.3333333   2.9047619   6.4761905 258.7333333   3.5466667   3.6190476
#>  [7] 159.8095238  17.7004762   0.3809524   3.3284286
#> attr(,"scale")
#>  [1]   0.4714045   1.6302136   1.7353207 124.9580780   0.5703995   0.7853535
#>  [7]  72.9393871   1.9347732   0.4856209   1.0629210
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 35
#> 
#> $lambda.min
#> [1] 0.4823412
#> 
#> $null.dev
#> [1] 33.25107
#> 
#> $Bias
#> [1] 2.459191
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 12.36405