Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 39.719407 36.236915 32.082025 27.562436 23.888438 20.932577 18.585167
#>   [8] 16.751898 15.351757 14.315230 13.582742 13.103299 12.833318 12.735607
#>  [15] 12.778488 12.929676 12.489128 12.112737 11.935970 11.935970 11.935970
#>  [22] 11.935970 11.935970 11.935970 11.907260 11.641877 11.426587 10.347170
#>  [29]  9.476795  9.476808  8.883489  8.227373  7.882153 10.903711 11.093575
#>  [36] 11.300325 10.892017 10.767655 10.626141 10.566867 10.553343 10.651518
#>  [43] 10.749003 10.839607 11.800382 11.873255 11.942482 12.363670 12.435892
#>  [50] 14.237631 14.316298 16.407659 16.415548 16.401226 16.355802 16.216404
#>  [57] 16.204910 16.469894 17.042463 18.504083 18.504258 20.253084 20.252600
#>  [64] 21.198825 22.344809 22.380069 22.380870 22.462103 25.248856 26.427551
#>  [71] 26.429762 26.431829 29.979543 29.979310 29.982112 33.923677 33.927620
#>  [78] 33.931458 33.996496 34.376597 34.400329 34.403864 34.407302 34.410646
#>  [85] 34.413900 34.434932 34.438074 34.595054 34.598175 34.601210 34.604163
#>  [92] 34.606976 34.684574 34.687887 34.833731 34.841877 34.845021 34.848075
#>  [99] 34.851042 34.853925
#> 
#> $cvse
#>   [1] 10.188308 10.322579 10.146738  9.267576  8.462191  7.722501  7.043512
#>   [8]  6.422736  5.859736  5.355761  4.913382  4.536027  4.227304  3.990071
#>  [15]  3.825322  3.730464  3.579167  3.436631  3.343213  3.343213  3.343213
#>  [22]  3.343213  3.343213  3.343213  3.345515  3.355394  3.370170  3.335712
#>  [29]  3.262751  3.262741  3.040008  2.820990  2.710334  3.356593  3.351390
#>  [36]  3.356015  3.313817  3.292331  3.265640  3.252808  3.267202  3.304238
#>  [43]  3.339367  3.374215  3.691167  3.719410  3.746038  3.856249  3.850867
#>  [50]  3.957342  3.936958  4.772522  4.787157  4.783899  4.786016  4.720940
#>  [57]  4.718210  4.849455  5.075242  5.319727  5.319742  6.235462  6.235142
#>  [64]  6.316375  6.329411  6.272568  6.272687  6.305464  6.874317  7.245026
#>  [71]  7.245726  7.246381  8.690187  8.691885  8.693170  9.990833  9.992353
#>  [78]  9.993834  9.990429 10.069443 10.073443 10.074805 10.076133 10.077429
#>  [85] 10.078692 10.076940 10.078134 10.059255 10.060375 10.061468 10.062535
#>  [92] 10.063582 10.056428 10.057359 10.126078 10.126197 10.127119 10.128023
#>  [99] 10.128909 10.129777
#> 
#> $fold
#>  [1]  9  2  6  2  7  3  1  3 10  8  1  5  8  6  7  5 10  9  4  1  4
#> 
#> $lambda
#>   [1] 5.711583902 5.326642263 4.967644402 4.632841795 4.320603764 4.029409531
#>   [7] 3.757840814 3.504574921 3.268378300 3.048100542 2.842668767 2.651082406
#>  [13] 2.472408325 2.305776279 2.150374675 2.005446619 1.870286228 1.744235196
#>  [19] 1.626679582 1.517046822 1.414802944 1.319449961 1.230523449 1.147590287
#>  [25] 1.070246543 0.998115508 0.930845863 0.868109967 0.809602261 0.755037778
#>  [31] 0.704150759 0.656693355 0.612434420 0.571158390 0.532664226 0.496764440
#>  [37] 0.463284180 0.432060376 0.402940953 0.375784081 0.350457492 0.326837830
#>  [43] 0.304810054 0.284266877 0.265108241 0.247240832 0.230577628 0.215037467
#>  [49] 0.200544662 0.187028623 0.174423520 0.162667958 0.151704682 0.141480294
#>  [55] 0.131944996 0.123052345 0.114759029 0.107024654 0.099811551 0.093084586
#>  [61] 0.086810997 0.080960226 0.075503778 0.070415076 0.065669335 0.061243442
#>  [67] 0.057115839 0.053266423 0.049676444 0.046328418 0.043206038 0.040294095
#>  [73] 0.037578408 0.035045749 0.032683783 0.030481005 0.028426688 0.026510824
#>  [79] 0.024724083 0.023057763 0.021503747 0.020054466 0.018702862 0.017442352
#>  [85] 0.016266796 0.015170468 0.014148029 0.013194500 0.012305234 0.011475903
#>  [91] 0.010702465 0.009981155 0.009308459 0.008681100 0.008096023 0.007550378
#>  [97] 0.007041508 0.006566934 0.006124344 0.005711584
#> 
#> $fit
#> $beta
#>             5.7116     5.3266     4.9676     4.6328    4.3206    4.0294
#> (Intercept)   20.4 22.2216444 24.0315698 25.7195123 27.293693 28.761779
#> am             0.0  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> carb           0.0  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> cyl            0.0 -0.2942656 -0.5866382 -0.8593058 -1.113597 -1.350749
#> disp           0.0  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> drat           0.0  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> gear           0.0  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> hp             0.0  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> qsec           0.0  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> vs             0.0  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> wt             0.0  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#>                3.7578    3.5046    3.2684   3.0481    2.8427    2.6511
#> (Intercept) 30.130922 31.407788 32.598598 33.70915 34.744858 35.710760
#> am           0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> cyl         -1.571918 -1.778181 -1.970543 -2.14994 -2.317246 -2.473277
#> disp         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#>                2.4724    2.3058   2.1504    2.0054    1.8703    1.7442
#> (Intercept) 36.611564 37.451657 38.23513 38.965801 39.517647 39.517647
#> am           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> cyl         -2.618791 -2.754499 -2.88106 -2.999091 -3.088235 -3.088235
#> disp         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#>                1.6267    1.5170    1.4148    1.3194    1.2305    1.1476
#> (Intercept) 39.517647 39.517647 39.517647 39.517647 39.517647 39.517647
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -3.088235 -3.088235 -3.088235 -3.088235 -3.088235 -3.088235
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.0702     0.9981    0.9308    0.8681    0.8096    0.7550
#> (Intercept) 39.517647 39.9045379 40.481055 40.658540 40.658304 40.658158
#> am           0.000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -3.088235 -2.6946537 -2.105196 -1.922919 -1.922785 -1.922702
#> disp         0.000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000 -0.8868229 -2.214079 -2.624256 -2.624442 -2.624557
#>                0.7042    0.6567   0.6124    0.5712    0.5327    0.4968
#> (Intercept) 40.658068 40.658013 40.65798 40.657957 40.657944 40.657936
#> am           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> cyl         -1.922651 -1.922620 -1.92260 -1.922588 -1.922581 -1.922576
#> disp         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> wt          -2.624628 -2.624672 -2.62470 -2.624716 -2.624727 -2.624733
#>                0.4633    0.4321    0.4029    0.3758    0.3505     0.3268
#> (Intercept) 40.657931 40.657928 40.657926 40.657925 40.657924 41.2027934
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> cyl         -1.922573 -1.922571 -1.922570 -1.922570 -1.922569 -1.9266055
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000 -0.1040407
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> wt          -2.624737 -2.624740 -2.624741 -2.624742 -2.624743 -2.6651021
#>                 0.3048     0.2843     0.2651     0.2472     0.2306     0.2150
#> (Intercept) 41.7426196 42.2425197 42.7085425 43.1431558 43.5484777 43.9244549
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl         -1.9303443 -1.9341075 -1.9376173 -1.9408904 -1.9439430 -1.9466459
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear        -0.2072071 -0.3026406 -0.3916062 -0.4745757 -0.5519533 -0.6237727
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.7054893 -2.7424254 -2.7768559 -2.8089659 -2.8389119 -2.8668881
#>                0.2005     0.1870     0.1744     0.1627     0.1517     0.1415
#> (Intercept) 44.277226 44.5909389 44.5918191 44.5918970 44.5917558 32.6924032
#> am           0.000000  0.0000000  0.0000000  0.0000000  0.0000000  1.8981259
#> carb         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl         -1.949306 -1.9515806 -1.9517453 -1.9518488 -1.9519125 -1.5016548
#> disp         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear        -0.691117 -0.7510361 -0.7511503 -0.7511301 -0.7510811 -0.8263983
#> hp           0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.4570643
#> vs           0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.892946 -2.9162599 -2.9160811 -2.9159282 -2.9158180 -2.7739707
#>                 0.1319     0.1231     0.1148     0.1070     0.0998     0.0931
#> (Intercept) 32.6873480 32.6829705 32.6791798 32.6758973 32.6730549 32.7619636
#> am           1.8985775  1.8989686  1.8993072  1.8996005  1.8998544  2.0406209
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl         -1.5013804 -1.5011428 -1.5009370 -1.5007588 -1.5006046 -1.3923060
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear        -0.8263110 -0.8262354 -0.8261699 -0.8261132 -0.8260641 -0.8686954
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.4572469  0.4574050  0.4575419  0.4576605  0.4577632  0.3851172
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.8168989
#> wt          -2.7741007 -2.7742133 -2.7743108 -2.7743952 -2.7744683 -2.6856233
#>                 0.0868     0.0810     0.0755     0.0704     0.0657     0.0612
#> (Intercept) 32.7661238 32.7698645 32.7732142 32.7762006 32.7788513 32.7811940
#> am           2.0404378  2.0402353  2.0400261  2.0398193  2.0396210  2.0394350
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl         -1.3925997 -1.3928558 -1.3930790 -1.3932733 -1.3934425 -1.3935898
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear        -0.8687901 -0.8688665 -0.8689293 -0.8689818 -0.8690262 -0.8690641
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.3849687  0.3848352  0.3847159  0.3846102  0.3845170  0.3844352
#> vs           0.8168337  0.8167747  0.8167195  0.8166662  0.8166141  0.8165631
#> wt          -2.6853841 -2.6851895 -2.6850318 -2.6849041 -2.6848006 -2.6847166
#>                 0.0571     0.0533       0.0497       0.0463      0.0432
#> (Intercept) 32.7832565 32.7850656 30.354495667 29.111381997 29.10759580
#> am           2.0392635  2.0391072  2.275849892  2.196162685  2.19652419
#> carb         0.0000000  0.0000000  0.000000000  0.000000000  0.00000000
#> cyl         -1.3937182 -1.3938300 -1.115787790 -1.072241371 -1.07193302
#> disp         0.0000000  0.0000000  0.000000000  0.000000000  0.00000000
#> drat         0.0000000  0.0000000  0.000000000  0.277215172  0.27717316
#> gear        -0.8690966 -0.8691244 -0.534045270 -0.571663846 -0.57145644
#> hp           0.0000000  0.0000000 -0.007564066 -0.007637451 -0.00764067
#> qsec         0.3843639  0.3843018  0.358366719  0.365154477  0.36523007
#> vs           0.8165134  0.8164653  1.052708701  1.061839690  1.06213533
#> wt          -2.6846480 -2.6845918 -2.424737174 -2.417246377 -2.41721473
#>                   0.0403     0.0376       0.0350       0.0327       0.0305
#> (Intercept) 29.104099258 29.1008705 29.097889144 29.095136301 29.092594420
#> am           2.196859256  2.1971696  2.197456884  2.197722690  2.197968532
#> carb         0.000000000  0.0000000  0.000000000  0.000000000  0.000000000
#> cyl         -1.071648200 -1.0713851 -1.071142122 -1.070917677 -1.070710360
#> disp         0.000000000  0.0000000  0.000000000  0.000000000  0.000000000
#> drat         0.277134463  0.2770988  0.277065975  0.277035722  0.277007851
#> gear        -0.571265124 -0.5710886 -0.570925622 -0.570775152 -0.570636178
#> hp          -0.007643647 -0.0076464 -0.007648947 -0.007651303 -0.007653482
#> qsec         0.365299746  0.3653640  0.365423150  0.365477693  0.365527961
#> vs           1.062409195  1.0626628  1.062897536  1.063114786  1.063315802
#> wt          -2.417184835 -2.4171567 -2.417130142 -2.417105234 -2.417081868
#>                   0.0284       0.0265       0.0247       0.0231       0.0215
#> (Intercept) 29.090247254 29.088079776 29.086078098 29.084229395 29.082521830
#> am           2.198195842  2.198405967  2.198600171  2.198779639  2.198945473
#> carb         0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> cyl         -1.070518857 -1.070341950 -1.070178518 -1.070027521 -1.069888004
#> disp         0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> drat         0.276982170  0.276958504  0.276936693  0.276916586  0.276898048
#> gear        -0.570507791 -0.570389157 -0.570279513 -0.570178161 -0.570084458
#> hp          -0.007655499 -0.007657365 -0.007659091 -0.007660689 -0.007662167
#> qsec         0.365574291  0.365616996  0.365656365  0.365692662  0.365726132
#> vs           1.063501767  1.063673786  1.063832891  1.063980041  1.064116128
#> wt          -2.417059976 -2.417039489 -2.417020336 -2.417002446 -2.416985751
#>                   0.0201       0.0187       0.0174       0.0163       0.0152
#> (Intercept) 29.080944485 29.079487288 29.078140956 29.076896930 29.075747322
#> am           2.199098700  2.199240275  2.199371084  2.199491947  2.199603624
#> carb         0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> cyl         -1.069759081 -1.069639940 -1.069529826 -1.069428049 -1.069333968
#> disp         0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> drat         0.276880953  0.276865186  0.276850641  0.276837222  0.276824837
#> gear        -0.569997815 -0.569917689 -0.569843580 -0.569775031 -0.569711618
#> hp          -0.007663535 -0.007664802 -0.007665974 -0.007667059 -0.007668063
#> qsec         0.365757000  0.365785473  0.365811742  0.365835981  0.365858351
#> vs           1.064241981  1.064358368  1.064466002  1.064565541  1.064657596
#> wt          -2.416970182 -2.416955673 -2.416942159 -2.416929581 -2.416917878
#>                   0.0141       0.0132       0.0123       0.0115       0.0107
#> (Intercept) 29.074684861 29.073702845 29.072795097 29.071955922 29.071180072
#> am           2.199706817  2.199802177  2.199890302  2.199971748  2.200047026
#> carb         0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> cyl         -1.069246994 -1.069166583 -1.069092235 -1.069023485 -1.068959909
#> disp         0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> drat         0.276813407  0.276802854  0.276793111  0.276784113  0.276775803
#> gear        -0.569652951 -0.569598670 -0.569548445 -0.569501970 -0.569458962
#> hp          -0.007668992 -0.007669852 -0.007670648 -0.007671386 -0.007672068
#> qsec         0.365878999  0.365898061  0.365915661  0.365931916  0.365946928
#> vs           1.064742732  1.064821472  1.064894298  1.064961658  1.065023964
#> wt          -2.416906995 -2.416896878 -2.416887477 -2.416878745 -2.416870635
#>                   0.0100       0.0093       0.0087       0.0081      0.0076
#> (Intercept) 29.070462707 29.069799362 29.069185922 29.068618590 29.06809386
#> am           2.200116608  2.200180930  2.200240395  2.200295372  2.20034621
#> carb         0.000000000  0.000000000  0.000000000  0.000000000  0.00000000
#> cyl         -1.068901113 -1.068846733 -1.068796434 -1.068749907 -1.06870687
#> disp         0.000000000  0.000000000  0.000000000  0.000000000  0.00000000
#> drat         0.276768126  0.276761034  0.276754480  0.276748424  0.27674283
#> gear        -0.569419161 -0.569382326 -0.569348234 -0.569316681 -0.56928748
#> hp          -0.007672699 -0.007673284 -0.007673825 -0.007674326 -0.00767479
#> qsec         0.365960796  0.365973609  0.365985448  0.365996389  0.36600650
#> vs           1.065081597  1.065134911  1.065184232  1.065229859  1.06527207
#> wt          -2.416863107 -2.416856120 -2.416849636 -2.416843621 -2.41683804
#>                  0.0070       0.0066        0.0061       0.0057
#> (Intercept) 29.06760851 29.067159542  2.906655e+01 29.202296245
#> am           2.20039321  2.200436677  2.200515e+00  2.198872051
#> carb         0.00000000  0.000000000  2.183562e-05  0.017280467
#> cyl         -1.06866705 -1.068630212 -1.068576e+00 -1.073128944
#> disp         0.00000000  0.000000000  0.000000e+00  0.000000000
#> drat         0.27673765  0.276732868  2.767075e-01  0.266813520
#> gear        -0.56926044 -0.569235421 -5.692045e-01 -0.583459077
#> hp          -0.00767522 -0.007675617 -7.676549e-03 -0.007879906
#> qsec         0.36601585  0.366024490  3.660383e-01  0.365645250
#> vs           1.06531113  1.065347263  1.065411e+00  1.062829383
#> wt          -2.41683287 -2.416828069 -2.416823e+00 -2.425432302
#> 
#> $iter
#>   [1]  0  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1  1  1  1  1
#>  [26] 63 69 15  1  1  1  1  1  1  1  1  1  1  1  1  1 10  8  8  8  8  8  7  7  6
#>  [51]  1  1  1 80  1  1  1  1  1 33  1  1  1  1  1  1  1  1 79 58  1  1  1  1  1
#>  [76]  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2 19
#> 
#> $lambda
#>   [1] 5.711583902 5.326642263 4.967644402 4.632841795 4.320603764 4.029409531
#>   [7] 3.757840814 3.504574921 3.268378300 3.048100542 2.842668767 2.651082406
#>  [13] 2.472408325 2.305776279 2.150374675 2.005446619 1.870286228 1.744235196
#>  [19] 1.626679582 1.517046822 1.414802944 1.319449961 1.230523449 1.147590287
#>  [25] 1.070246543 0.998115508 0.930845863 0.868109967 0.809602261 0.755037778
#>  [31] 0.704150759 0.656693355 0.612434420 0.571158390 0.532664226 0.496764440
#>  [37] 0.463284180 0.432060376 0.402940953 0.375784081 0.350457492 0.326837830
#>  [43] 0.304810054 0.284266877 0.265108241 0.247240832 0.230577628 0.215037467
#>  [49] 0.200544662 0.187028623 0.174423520 0.162667958 0.151704682 0.141480294
#>  [55] 0.131944996 0.123052345 0.114759029 0.107024654 0.099811551 0.093084586
#>  [61] 0.086810997 0.080960226 0.075503778 0.070415076 0.065669335 0.061243442
#>  [67] 0.057115839 0.053266423 0.049676444 0.046328418 0.043206038 0.040294095
#>  [73] 0.037578408 0.035045749 0.032683783 0.030481005 0.028426688 0.026510824
#>  [79] 0.024724083 0.023057763 0.021503747 0.020054466 0.018702862 0.017442352
#>  [85] 0.016266796 0.015170468 0.014148029 0.013194500 0.012305234 0.011475903
#>  [91] 0.010702465 0.009981155 0.009308459 0.008681100 0.008096023 0.007550378
#>  [97] 0.007041508 0.006566934 0.006124344 0.005711584
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 25
#> 
#> $loss
#>   [1] 828.46000 705.15179 594.85538 502.96811 426.81968 364.10598 312.84013
#>   [8] 271.31003 238.04146 211.76613 191.39388 175.98864 164.74753 156.98275
#>  [15] 152.10590 149.61434 149.04824 149.04824 149.04824 149.04824 149.04824
#>  [22] 149.04824 149.04824 149.04824 149.04824 115.43928  90.66655  89.20154
#>  [29]  89.20154  89.20154  89.20154  89.20154  89.20154  89.20154  89.20154
#>  [36]  89.20154  89.20154  89.20154  89.20154  89.20154  89.20154  88.13586
#>  [43]  87.23578  86.54205  86.01554  85.62905  85.35953  85.18788  85.09560
#>  [50]  85.06937  85.06937  85.06937  85.06937  80.26323  80.26322  80.26322
#>  [57]  80.26321  80.26321  80.26320  79.42439  79.42438  79.42437  79.42436
#>  [64]  79.42436  79.42436  79.42435  79.42435  79.42435  78.82374  78.66848
#>  [71]  78.66847  78.66846  78.66846  78.66845  78.66844  78.66844  78.66844
#>  [78]  78.66843  78.66843  78.66843  78.66842  78.66842  78.66842  78.66842
#>  [85]  78.66842  78.66842  78.66842  78.66842  78.66842  78.66842  78.66842
#>  [92]  78.66841  78.66841  78.66841  78.66841  78.66841  78.66841  78.66841
#>  [99]  78.66840  78.66476
#> 
#> $linear.predictors
#>    5.7116   5.3266   4.9676   4.6328   4.3206   4.0294   3.7578   3.5046
#> 1    20.4 20.45605 20.51174 20.56368 20.61211 20.65729 20.69941 20.73870
#> 2    20.4 21.04458 21.68502 22.28229 22.83931 23.35878 23.84325 24.29506
#> 3    20.4 20.45605 20.51174 20.56368 20.61211 20.65729 20.69941 20.73870
#> 4    20.4 21.04458 21.68502 22.28229 22.83931 23.35878 23.84325 24.29506
#> 5    20.4 21.04458 21.68502 22.28229 22.83931 23.35878 23.84325 24.29506
#> 6    20.4 20.45605 20.51174 20.56368 20.61211 20.65729 20.69941 20.73870
#> 7    20.4 19.86752 19.33846 18.84507 18.38492 17.95579 17.55558 17.18234
#> 8    20.4 19.86752 19.33846 18.84507 18.38492 17.95579 17.55558 17.18234
#> 9    20.4 19.86752 19.33846 18.84507 18.38492 17.95579 17.55558 17.18234
#> 10   20.4 19.86752 19.33846 18.84507 18.38492 17.95579 17.55558 17.18234
#> 11   20.4 19.86752 19.33846 18.84507 18.38492 17.95579 17.55558 17.18234
#> 12   20.4 21.04458 21.68502 22.28229 22.83931 23.35878 23.84325 24.29506
#> 13   20.4 21.04458 21.68502 22.28229 22.83931 23.35878 23.84325 24.29506
#> 14   20.4 19.86752 19.33846 18.84507 18.38492 17.95579 17.55558 17.18234
#> 15   20.4 19.86752 19.33846 18.84507 18.38492 17.95579 17.55558 17.18234
#> 16   20.4 19.86752 19.33846 18.84507 18.38492 17.95579 17.55558 17.18234
#> 17   20.4 21.04458 21.68502 22.28229 22.83931 23.35878 23.84325 24.29506
#> 18   20.4 21.04458 21.68502 22.28229 22.83931 23.35878 23.84325 24.29506
#> 19   20.4 21.04458 21.68502 22.28229 22.83931 23.35878 23.84325 24.29506
#> 20   20.4 19.86752 19.33846 18.84507 18.38492 17.95579 17.55558 17.18234
#> 21   20.4 19.86752 19.33846 18.84507 18.38492 17.95579 17.55558 17.18234
#>      3.2684   3.0481   2.8427   2.6511   2.4724   2.3058   2.1504   2.0054
#> 1  20.77534 20.80951 20.84138 20.87110 20.89882 20.92467 20.94877 20.97126
#> 2  24.71643 25.10939 25.47587 25.81765 26.13640 26.43366 26.71089 26.96944
#> 3  20.77534 20.80951 20.84138 20.87110 20.89882 20.92467 20.94877 20.97126
#> 4  24.71643 25.10939 25.47587 25.81765 26.13640 26.43366 26.71089 26.96944
#> 5  24.71643 25.10939 25.47587 25.81765 26.13640 26.43366 26.71089 26.96944
#> 6  20.77534 20.80951 20.84138 20.87110 20.89882 20.92467 20.94877 20.97126
#> 7  16.83426 16.50963 16.20689 15.92455 15.66124 15.41567 15.18665 14.97307
#> 8  16.83426 16.50963 16.20689 15.92455 15.66124 15.41567 15.18665 14.97307
#> 9  16.83426 16.50963 16.20689 15.92455 15.66124 15.41567 15.18665 14.97307
#> 10 16.83426 16.50963 16.20689 15.92455 15.66124 15.41567 15.18665 14.97307
#> 11 16.83426 16.50963 16.20689 15.92455 15.66124 15.41567 15.18665 14.97307
#> 12 24.71643 25.10939 25.47587 25.81765 26.13640 26.43366 26.71089 26.96944
#> 13 24.71643 25.10939 25.47587 25.81765 26.13640 26.43366 26.71089 26.96944
#> 14 16.83426 16.50963 16.20689 15.92455 15.66124 15.41567 15.18665 14.97307
#> 15 16.83426 16.50963 16.20689 15.92455 15.66124 15.41567 15.18665 14.97307
#> 16 16.83426 16.50963 16.20689 15.92455 15.66124 15.41567 15.18665 14.97307
#> 17 24.71643 25.10939 25.47587 25.81765 26.13640 26.43366 26.71089 26.96944
#> 18 24.71643 25.10939 25.47587 25.81765 26.13640 26.43366 26.71089 26.96944
#> 19 24.71643 25.10939 25.47587 25.81765 26.13640 26.43366 26.71089 26.96944
#> 20 16.83426 16.50963 16.20689 15.92455 15.66124 15.41567 15.18665 14.97307
#> 21 16.83426 16.50963 16.20689 15.92455 15.66124 15.41567 15.18665 14.97307
#>      1.8703   1.7442   1.6267   1.5170   1.4148   1.3194   1.2305   1.1476
#> 1  20.98824 20.98824 20.98824 20.98824 20.98824 20.98824 20.98824 20.98824
#> 2  27.16471 27.16471 27.16471 27.16471 27.16471 27.16471 27.16471 27.16471
#> 3  20.98824 20.98824 20.98824 20.98824 20.98824 20.98824 20.98824 20.98824
#> 4  27.16471 27.16471 27.16471 27.16471 27.16471 27.16471 27.16471 27.16471
#> 5  27.16471 27.16471 27.16471 27.16471 27.16471 27.16471 27.16471 27.16471
#> 6  20.98824 20.98824 20.98824 20.98824 20.98824 20.98824 20.98824 20.98824
#> 7  14.81176 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176
#> 8  14.81176 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176
#> 9  14.81176 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176
#> 10 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176
#> 11 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176
#> 12 27.16471 27.16471 27.16471 27.16471 27.16471 27.16471 27.16471 27.16471
#> 13 27.16471 27.16471 27.16471 27.16471 27.16471 27.16471 27.16471 27.16471
#> 14 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176
#> 15 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176
#> 16 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176
#> 17 27.16471 27.16471 27.16471 27.16471 27.16471 27.16471 27.16471 27.16471
#> 18 27.16471 27.16471 27.16471 27.16471 27.16471 27.16471 27.16471 27.16471
#> 19 27.16471 27.16471 27.16471 27.16471 27.16471 27.16471 27.16471 27.16471
#> 20 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176
#> 21 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176 14.81176
#>      1.0702   0.9981   0.9308   0.8681   0.8096   0.7550   0.7042   0.6567
#> 1  20.98824 21.41314 22.04899 22.24548 22.24556 22.24560 22.24563 22.24565
#> 2  27.16471 27.06849 26.92361 26.87859 26.87846 26.87838 26.87833 26.87829
#> 3  20.98824 20.88548 20.73162 20.68404 20.68401 20.68399 20.68398 20.68397
#> 4  27.16471 26.29696 24.99736 24.59549 24.59519 24.59501 24.59490 24.59483
#> 5  27.16471 26.33243 25.08592 24.70046 24.70017 24.69999 24.69988 24.69982
#> 6  20.98824 20.68595 20.23345 20.09359 20.09351 20.09347 20.09344 20.09342
#> 7  14.81176 14.73794 14.62819 14.59447 14.59454 14.59459 14.59462 14.59464
#> 8  14.81176 15.03946 15.38097 15.48671 15.48685 15.48694 15.48699 15.48703
#> 9  14.81176 14.99512 15.27027 15.35550 15.35563 15.35571 15.35576 15.35579
#> 10 14.81176 13.53718 11.63032 11.04122 11.04105 11.04094 11.04087 11.04083
#> 11 14.81176 13.60724 11.80524 11.24854 11.24838 11.24828 11.24822 11.24818
#> 12 27.16471 27.69370 28.48453 28.72869 28.72869 28.72869 28.72869 28.72869
#> 13 27.16471 27.49860 27.99744 28.15135 28.15131 28.15129 28.15127 28.15126
#> 14 14.81176 15.22569 15.84593 16.03781 16.03799 16.03810 16.03817 16.03821
#> 15 14.81176 15.30107 16.03413 16.26087 16.26106 16.26119 16.26126 16.26131
#> 16 14.81176 14.94191 15.13742 15.19805 15.19817 15.19824 15.19829 15.19831
#> 17 27.16471 27.40992 27.77603 27.88893 27.88887 27.88883 27.88881 27.88879
#> 18 27.16471 27.22812 27.32214 27.35096 27.35086 27.35080 27.35076 27.35074
#> 19 27.16471 27.78416 28.71037 28.99636 28.99638 28.99639 28.99640 28.99641
#> 20 14.81176 15.53608 16.62086 16.95630 16.95654 16.95669 16.95679 16.95684
#> 21 14.81176 15.18135 15.73523 15.90660 15.90677 15.90687 15.90694 15.90698
#>      0.6124   0.5712   0.5327   0.4968   0.4633   0.4321   0.4029   0.3758
#> 1  22.24567 22.24567 22.24568 22.24568 22.24568 22.24568 22.24568 22.24568
#> 2  26.87828 26.87826 26.87826 26.87825 26.87825 26.87825 26.87825 26.87824
#> 3  20.68397 20.68397 20.68396 20.68396 20.68396 20.68396 20.68396 20.68396
#> 4  24.59479 24.59476 24.59474 24.59473 24.59473 24.59472 24.59472 24.59472
#> 5  24.69977 24.69975 24.69973 24.69972 24.69972 24.69971 24.69971 24.69971
#> 6  20.09341 20.09340 20.09340 20.09340 20.09340 20.09340 20.09340 20.09339
#> 7  14.59465 14.59466 14.59466 14.59466 14.59467 14.59467 14.59467 14.59467
#> 8  15.48705 15.48706 15.48707 15.48707 15.48708 15.48708 15.48708 15.48708
#> 9  15.35581 15.35583 15.35583 15.35584 15.35584 15.35584 15.35584 15.35584
#> 10 11.04081 11.04079 11.04078 11.04078 11.04077 11.04077 11.04077 11.04077
#> 11 11.24816 11.24814 11.24814 11.24813 11.24813 11.24812 11.24812 11.24812
#> 12 28.72869 28.72869 28.72869 28.72869 28.72869 28.72869 28.72869 28.72869
#> 13 28.15125 28.15125 28.15125 28.15125 28.15125 28.15125 28.15125 28.15124
#> 14 16.03824 16.03825 16.03826 16.03827 16.03827 16.03827 16.03828 16.03828
#> 15 16.26133 16.26135 16.26136 16.26137 16.26137 16.26138 16.26138 16.26138
#> 16 15.19833 15.19834 15.19835 15.19835 15.19836 15.19836 15.19836 15.19836
#> 17 27.88878 27.88878 27.88878 27.88877 27.88877 27.88877 27.88877 27.88877
#> 18 27.35072 27.35071 27.35071 27.35070 27.35070 27.35070 27.35070 27.35070
#> 19 28.99641 28.99641 28.99641 28.99641 28.99641 28.99641 28.99641 28.99641
#> 20 16.95688 16.95690 16.95692 16.95692 16.95693 16.95693 16.95693 16.95694
#> 21 15.90700 15.90702 15.90703 15.90703 15.90703 15.90704 15.90704 15.90704
#>      0.3505   0.3268   0.3048   0.2843   0.2651   0.2472   0.2306   0.2150
#> 1  22.24568 22.24443 22.24334 22.24216 22.24105 22.24002 22.23906 22.23824
#> 2  26.87824 26.89717 26.91568 26.93310 26.94934 26.96449 26.97862 26.99160
#> 3  20.68396 20.76274 20.84078 20.91306 20.98043 21.04326 21.10186 21.15622
#> 4  24.59472 24.57853 24.56190 24.54719 24.53348 24.52069 24.50876 24.49741
#> 5  24.69971 24.68514 24.67012 24.65689 24.64455 24.63305 24.62232 24.61208
#> 6  20.09339 20.05905 20.02484 19.99337 19.96403 19.93667 19.91115 19.88739
#> 7  14.59467 14.63086 14.66690 14.70007 14.73098 14.75981 14.78670 14.81173
#> 8  15.48708 15.53700 15.58677 15.63249 15.67511 15.71486 15.75193 15.78648
#> 9  15.35584 15.40374 15.45149 15.49537 15.53627 15.57441 15.60999 15.64313
#> 10 11.04077 11.02231 11.00367 10.98682 10.97112 10.95647 10.94282 10.92997
#> 11 11.24812 11.23286 11.21740 11.20347 11.19049 11.17838 11.16709 11.15645
#> 12 28.72869 28.77607 28.82305 28.86651 28.90703 28.94481 28.98005 29.01276
#> 13 28.15124 28.18975 28.22784 28.26318 28.29612 28.32684 28.35549 28.38204
#> 14 16.03828 16.09667 16.15492 16.20840 16.25825 16.30475 16.34810 16.38852
#> 15 16.26138 16.32320 16.38489 16.44151 16.49429 16.54351 16.58941 16.63221
#> 16 15.19836 15.24384 15.28917 15.33082 15.36966 15.40588 15.43965 15.47112
#> 17 27.88877 27.92324 27.95729 27.98893 28.01843 28.04594 28.07160 28.09535
#> 18 27.35070 27.27285 27.19546 27.12410 27.05757 26.99553 26.93767 26.88387
#> 19 28.99641 28.94387 28.89180 28.84360 28.79866 28.75675 28.71767 28.68141
#> 20 16.95694 16.82137 16.68743 16.56297 16.44694 16.33873 16.23782 16.14439
#> 21 15.90704 15.75533 15.60523 15.46600 15.33620 15.21515 15.10225 14.99763
#>      0.2005   0.1870   0.1744   0.1627   0.1517   0.1415   0.1319   0.1231
#> 1  22.23740 22.23671 22.23661 22.23655 22.23651 22.53048 22.53054 22.53059
#> 2  27.00390 27.01475 27.01493 27.01503 27.01508 27.34867 27.34861 27.34856
#> 3  21.20722 21.25257 21.25270 21.25270 21.25268 21.17029 21.17028 21.17027
#> 4  24.48704 24.47760 24.47794 24.47817 24.47832 23.67251 23.67214 23.67182
#> 5  24.60275 24.59425 24.59458 24.59481 24.59495 25.10895 25.10912 25.10926
#> 6  19.86519 19.84538 19.84543 19.84549 19.84554 19.19870 19.19853 19.19839
#> 7  14.83513 14.85601 14.85596 14.85589 14.85583 14.86283 14.86288 14.86292
#> 8  15.81874 15.84754 15.84742 15.84730 15.84721 15.89739 15.89752 15.89763
#> 9  15.67409 15.70172 15.70162 15.70151 15.70142 15.94152 15.94171 15.94189
#> 10 10.91809 10.90739 10.90758 10.90772 10.90782 11.29884 11.29879 11.29875
#> 11 11.14663 11.13778 11.13795 11.13808 11.13817 11.33516 11.33504 11.33495
#> 12 29.04343 29.07071 29.07077 29.07076 29.07074 29.26318 29.26320 29.26321
#> 13 28.40698 28.42914 28.42923 28.42925 28.42926 29.28366 29.28390 29.28410
#> 14 16.42626 16.45995 16.45980 16.45965 16.45953 16.14627 16.14629 16.14631
#> 15 16.67216 16.70783 16.70767 16.70750 16.70738 16.57859 16.57871 16.57881
#> 16 15.50051 15.52675 15.52665 15.52655 15.52647 14.59128 14.59100 14.59075
#> 17 28.11768 28.13751 28.13762 28.13766 28.13767 28.54920 28.54924 28.54928
#> 18 26.83351 26.78864 26.78867 26.78876 26.78885 26.14859 26.14830 26.14804
#> 19 28.64739 28.61713 28.61706 28.61705 28.61707 27.97929 27.97911 27.97895
#> 20 16.05655 15.97857 15.97813 15.97796 15.97791 16.27924 16.27951 16.27974
#> 21 14.89937 14.81207 14.81170 14.81159 14.81158 15.21536 15.21559 15.21579
#>      0.1148   0.1070   0.0998   0.0931   0.0868   0.0810   0.0755   0.0704
#> 1  22.53063 22.53067 22.53070 22.27666 22.27668 22.27669 22.27669 22.27669
#> 2  27.34851 27.34847 27.34844 27.51186 27.51201 27.51213 27.51222 27.51228
#> 3  21.17025 21.17024 21.17024 21.47134 21.47127 21.47122 21.47117 21.47114
#> 4  23.67154 23.67130 23.67109 23.67006 23.67040 23.67070 23.67097 23.67121
#> 5  25.10938 25.10949 25.10958 24.89433 24.89422 24.89413 24.89405 24.89397
#> 6  19.19827 19.19816 19.19807 19.55935 19.55940 19.55947 19.55954 19.55960
#> 7  14.86296 14.86299 14.86302 14.78798 14.78790 14.78783 14.78777 14.78773
#> 8  15.89773 15.89782 15.89789 15.77812 15.77792 15.77776 15.77763 15.77752
#> 9  15.94203 15.94216 15.94227 15.79788 15.79764 15.79744 15.79726 15.79712
#> 10 11.29871 11.29868 11.29865 11.31340 11.31357 11.31371 11.31382 11.31390
#> 11 11.33486 11.33479 11.33473 11.37151 11.37173 11.37191 11.37205 11.37217
#> 12 29.26322 29.26323 29.26324 29.37057 29.37056 29.37055 29.37054 29.37053
#> 13 29.28428 29.28444 29.28457 29.31119 29.31103 29.31088 29.31074 29.31061
#> 14 16.14633 16.14635 16.14636 16.06096 16.06083 16.06072 16.06064 16.06058
#> 15 16.57889 16.57897 16.57903 16.45484 16.45462 16.45444 16.45430 16.45418
#> 16 14.59054 14.59036 14.59020 14.63929 14.63945 14.63960 14.63975 14.63988
#> 17 28.54931 28.54934 28.54936 28.65751 28.65753 28.65753 28.65752 28.65751
#> 18 26.14781 26.14762 26.14745 25.57411 25.57447 25.57479 25.57507 25.57531
#> 19 27.97882 27.97870 27.97860 28.15191 28.15203 28.15214 28.15224 28.15234
#> 20 16.27993 16.28011 16.28025 16.39143 16.39119 16.39098 16.39079 16.39063
#> 21 15.21596 15.21611 15.21624 15.35569 15.35554 15.35539 15.35525 15.35513
#>      0.0657   0.0612   0.0571   0.0533   0.0497   0.0463   0.0432   0.0403
#> 1  22.27668 22.27668 22.27668 22.27667 22.51330 22.50572 22.50578 22.50584
#> 2  27.51234 27.51237 27.51240 27.51242 27.42408 27.33827 27.33823 27.33818
#> 3  21.47111 21.47108 21.47106 21.47104 21.44941 21.36564 21.36571 21.36577
#> 4  23.67142 23.67160 23.67176 23.67190 23.77132 23.73908 23.73891 23.73875
#> 5  24.89391 24.89385 24.89380 24.89376 24.65796 24.70644 24.70637 24.70631
#> 6  19.55965 19.55970 19.55974 19.55978 19.86293 19.96739 19.96752 19.96763
#> 7  14.78769 14.78766 14.78764 14.78762 14.83143 14.81026 14.81030 14.81034
#> 8  15.77743 15.77735 15.77729 15.77724 15.72751 15.70516 15.70520 15.70524
#> 9  15.79699 15.79689 15.79680 15.79673 15.74962 15.73036 15.73043 15.73050
#> 10 11.31397 11.31402 11.31405 11.31409 11.43410 11.40396 11.40396 11.40397
#> 11 11.37226 11.37233 11.37240 11.37245 11.36885 11.39806 11.39797 11.39789
#> 12 29.37051 29.37050 29.37049 29.37047 29.41139 29.62209 29.62211 29.62212
#> 13 29.31049 29.31038 29.31029 29.31020 29.27416 29.29810 29.29821 29.29832
#> 14 16.06054 16.06050 16.06048 16.06046 16.20202 16.08940 16.08950 16.08958
#> 15 16.45409 16.45401 16.45395 16.45390 16.56222 16.56000 16.56010 16.56020
#> 16 14.64000 14.64012 14.64022 14.64031 14.18430 14.32610 14.32574 14.32542
#> 17 28.65749 28.65748 28.65746 28.65744 28.66576 28.64478 28.64482 28.64485
#> 18 25.57553 25.57572 25.57589 25.57605 25.60442 25.61849 25.61818 25.61790
#> 19 28.15242 28.15249 28.15255 28.15261 28.08271 27.91799 27.91793 27.91788
#> 20 16.39048 16.39035 16.39023 16.39013 16.54680 16.65692 16.65717 16.65740
#> 21 15.35501 15.35491 15.35481 15.35472 15.07570 14.99578 14.99584 14.99591
#>      0.0376   0.0350   0.0327   0.0305   0.0284   0.0265   0.0247   0.0231
#> 1  22.50590 22.50595 22.50600 22.50604 22.50608 22.50612 22.50615 22.50618
#> 2  27.33814 27.33811 27.33807 27.33804 27.33802 27.33799 27.33797 27.33795
#> 3  21.36583 21.36589 21.36594 21.36599 21.36603 21.36607 21.36611 21.36614
#> 4  23.73861 23.73847 23.73835 23.73824 23.73813 23.73803 23.73794 23.73786
#> 5  24.70625 24.70619 24.70614 24.70610 24.70605 24.70601 24.70597 24.70594
#> 6  19.96773 19.96783 19.96791 19.96800 19.96807 19.96814 19.96821 19.96827
#> 7  14.81037 14.81040 14.81043 14.81045 14.81048 14.81050 14.81052 14.81054
#> 8  15.70528 15.70531 15.70534 15.70537 15.70539 15.70542 15.70544 15.70546
#> 9  15.73056 15.73062 15.73068 15.73072 15.73077 15.73081 15.73085 15.73088
#> 10 11.40397 11.40398 11.40398 11.40398 11.40399 11.40399 11.40400 11.40400
#> 11 11.39782 11.39775 11.39769 11.39763 11.39758 11.39753 11.39749 11.39745
#> 12 29.62213 29.62213 29.62214 29.62215 29.62216 29.62217 29.62217 29.62218
#> 13 29.29841 29.29850 29.29858 29.29865 29.29872 29.29878 29.29884 29.29889
#> 14 16.08966 16.08973 16.08979 16.08985 16.08991 16.08996 16.09000 16.09005
#> 15 16.56029 16.56037 16.56044 16.56051 16.56058 16.56063 16.56069 16.56074
#> 16 14.32511 14.32483 14.32457 14.32433 14.32411 14.32390 14.32371 14.32354
#> 17 28.64489 28.64492 28.64495 28.64498 28.64501 28.64503 28.64505 28.64507
#> 18 25.61765 25.61741 25.61718 25.61698 25.61679 25.61662 25.61646 25.61631
#> 19 27.91783 27.91779 27.91775 27.91771 27.91768 27.91764 27.91761 27.91759
#> 20 16.65761 16.65781 16.65799 16.65816 16.65831 16.65846 16.65859 16.65871
#> 21 14.99597 14.99602 14.99607 14.99612 14.99616 14.99620 14.99623 14.99627
#>      0.0215   0.0201   0.0187   0.0174   0.0163   0.0152   0.0141   0.0132
#> 1  22.50621 22.50624 22.50627 22.50629 22.50631 22.50633 22.50635 22.50637
#> 2  27.33793 27.33791 27.33789 27.33787 27.33786 27.33785 27.33783 27.33782
#> 3  21.36617 21.36620 21.36623 21.36625 21.36627 21.36630 21.36631 21.36633
#> 4  23.73779 23.73772 23.73765 23.73759 23.73754 23.73749 23.73744 23.73740
#> 5  24.70591 24.70588 24.70585 24.70582 24.70580 24.70577 24.70575 24.70573
#> 6  19.96832 19.96837 19.96842 19.96847 19.96851 19.96855 19.96858 19.96861
#> 7  14.81055 14.81057 14.81058 14.81060 14.81061 14.81062 14.81063 14.81064
#> 8  15.70547 15.70549 15.70551 15.70552 15.70553 15.70554 15.70556 15.70557
#> 9  15.73092 15.73095 15.73097 15.73100 15.73102 15.73104 15.73106 15.73108
#> 10 11.40400 11.40400 11.40401 11.40401 11.40401 11.40402 11.40402 11.40402
#> 11 11.39741 11.39737 11.39734 11.39731 11.39728 11.39726 11.39723 11.39721
#> 12 29.62218 29.62219 29.62219 29.62220 29.62220 29.62221 29.62221 29.62221
#> 13 29.29894 29.29899 29.29903 29.29907 29.29911 29.29914 29.29917 29.29920
#> 14 16.09009 16.09013 16.09016 16.09019 16.09022 16.09025 16.09027 16.09030
#> 15 16.56078 16.56083 16.56087 16.56090 16.56094 16.56097 16.56099 16.56102
#> 16 14.32338 14.32323 14.32309 14.32296 14.32284 14.32273 14.32263 14.32254
#> 17 28.64509 28.64511 28.64512 28.64514 28.64515 28.64517 28.64518 28.64519
#> 18 25.61617 25.61604 25.61593 25.61582 25.61572 25.61563 25.61554 25.61546
#> 19 27.91756 27.91754 27.91752 27.91750 27.91748 27.91746 27.91745 27.91743
#> 20 16.65882 16.65893 16.65902 16.65911 16.65919 16.65927 16.65934 16.65940
#> 21 14.99630 14.99632 14.99635 14.99637 14.99640 14.99642 14.99643 14.99645
#>      0.0123   0.0115   0.0107   0.0100   0.0093   0.0087   0.0081   0.0076
#> 1  22.50639 22.50640 22.50642 22.50643 22.50644 22.50645 22.50646 22.50647
#> 2  27.33781 27.33780 27.33779 27.33778 27.33778 27.33777 27.33776 27.33776
#> 3  21.36635 21.36636 21.36638 21.36639 21.36640 21.36642 21.36643 21.36644
#> 4  23.73736 23.73732 23.73728 23.73725 23.73722 23.73720 23.73717 23.73715
#> 5  24.70572 24.70570 24.70569 24.70567 24.70566 24.70565 24.70563 24.70562
#> 6  19.96865 19.96867 19.96870 19.96872 19.96875 19.96877 19.96879 19.96880
#> 7  14.81065 14.81066 14.81067 14.81067 14.81068 14.81069 14.81069 14.81070
#> 8  15.70558 15.70558 15.70559 15.70560 15.70561 15.70561 15.70562 15.70562
#> 9  15.73110 15.73111 15.73113 15.73114 15.73115 15.73116 15.73117 15.73118
#> 10 11.40402 11.40402 11.40402 11.40403 11.40403 11.40403 11.40403 11.40403
#> 11 11.39719 11.39717 11.39716 11.39714 11.39713 11.39711 11.39710 11.39709
#> 12 29.62222 29.62222 29.62222 29.62223 29.62223 29.62223 29.62223 29.62223
#> 13 29.29922 29.29925 29.29927 29.29929 29.29931 29.29933 29.29934 29.29936
#> 14 16.09032 16.09034 16.09035 16.09037 16.09039 16.09040 16.09042 16.09043
#> 15 16.56105 16.56107 16.56109 16.56111 16.56113 16.56114 16.56116 16.56117
#> 16 14.32245 14.32237 14.32230 14.32223 14.32217 14.32211 14.32206 14.32201
#> 17 28.64520 28.64521 28.64522 28.64522 28.64523 28.64524 28.64524 28.64525
#> 18 25.61539 25.61532 25.61526 25.61520 25.61515 25.61510 25.61506 25.61502
#> 19 27.91742 27.91741 27.91740 27.91739 27.91738 27.91737 27.91736 27.91735
#> 20 16.65946 16.65952 16.65957 16.65962 16.65966 16.65970 16.65974 16.65978
#> 21 14.99647 14.99648 14.99649 14.99651 14.99652 14.99653 14.99654 14.99655
#>      0.0070   0.0066   0.0061   0.0057
#> 1  22.50648 22.50649 22.50654 22.53535
#> 2  27.33775 27.33774 27.33773 27.32698
#> 3  21.36644 21.36645 21.36645 21.35863
#> 4  23.73713 23.73711 23.73709 23.74510
#> 5  24.70561 24.70561 24.70559 24.70382
#> 6  19.96882 19.96884 19.96889 19.98614
#> 7  14.81070 14.81071 14.81071 14.81018
#> 8  15.70563 15.70563 15.70564 15.70796
#> 9  15.73119 15.73120 15.73121 15.73294
#> 10 11.40403 11.40403 11.40404 11.40252
#> 11 11.39708 11.39707 11.39705 11.39104
#> 12 29.62224 29.62224 29.62225 29.63252
#> 13 29.29937 29.29939 29.29940 29.29436
#> 14 16.09044 16.09045 16.09046 16.08678
#> 15 16.56118 16.56120 16.56120 16.55423
#> 16 14.32196 14.32192 14.32183 14.32158
#> 17 28.64526 28.64526 28.64526 28.64094
#> 18 25.61498 25.61494 25.61487 25.60668
#> 19 27.91734 27.91734 27.91733 27.91393
#> 20 16.65981 16.65984 16.65985 16.62686
#> 21 14.99655 14.99656 14.99662 15.02146
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 22.8 21.4 24.4 22.8 19.2 16.4 17.3 15.2 10.4 14.7 30.4 33.9 15.5 15.2
#> [16] 13.3 27.3 26.0 30.4 15.8 15.0
#> 
#> $X
#>               am       carb        cyl       disp        drat       gear
#>  [1,]  1.1547005  0.7476858 -0.1034175 -0.5710005  0.45674893  0.3175003
#>  [2,]  1.1547005 -1.1364823 -1.1893017 -1.0120647  0.36574583  0.3175003
#>  [3,] -0.8660254 -1.1364823 -0.1034175  0.2602360 -1.03570204 -1.0160010
#>  [4,] -0.8660254 -0.5084263 -1.1893017 -0.6838111  0.07453588  0.3175003
#>  [5,] -0.8660254 -0.5084263 -1.1893017 -0.7338549  0.49315018  0.3175003
#>  [6,] -0.8660254  0.7476858 -0.1034175 -0.5065372  0.49315018  0.3175003
#>  [7,] -0.8660254  0.1196297  0.9824666  0.4112157 -1.05390266 -1.0160010
#>  [8,] -0.8660254  0.1196297  0.9824666  0.4112157 -1.05390266 -1.0160010
#>  [9,] -0.8660254  0.1196297  0.9824666  0.4112157 -1.05390266 -1.0160010
#> [10,] -0.8660254  0.7476858  0.9824666  1.9736009 -1.18130702 -1.0160010
#> [11,] -0.8660254  0.7476858  0.9824666  1.8039608 -0.76269272 -1.0160010
#> [12,]  1.1547005 -0.5084263 -1.1893017 -1.2860334  2.33141297  0.3175003
#> [13,]  1.1547005 -1.1364823 -1.1893017 -1.3250507  1.03916883  0.3175003
#> [14,] -0.8660254 -0.5084263  0.9824666  0.7691562 -1.61812194 -1.0160010
#> [15,] -0.8660254 -0.5084263  0.9824666  0.6504082 -0.90829769 -1.0160010
#> [16,] -0.8660254  0.7476858  0.9824666  1.0405804  0.14733837 -1.0160010
#> [17,]  1.1547005 -1.1364823 -1.1893017 -1.2580428  0.78436012  0.3175003
#> [18,]  1.1547005 -0.5084263 -1.1893017 -0.9077360  1.42138188  1.6510017
#> [19,]  1.1547005 -0.5084263 -1.1893017 -1.1214826  0.22014085  1.6510017
#> [20,]  1.1547005  0.7476858  0.9824666  1.0490624  1.03916883  1.6510017
#> [21,]  1.1547005  3.2599099  0.9824666  0.6249622 -0.19847345  1.6510017
#>                hp        qsec         vs          wt
#>  [1,] -0.51154591 -0.71360293 -0.8660254 -0.54170473
#>  [2,] -0.74010898  0.44162017  1.1547005 -0.83001593
#>  [3,] -0.51154591  0.88759002  1.1547005  0.03011250
#>  [4,] -1.15690046  1.18848534  1.1547005  0.00608657
#>  [5,] -0.71321921  2.74669325  1.1547005 -0.03235492
#>  [6,] -0.33676239  0.27505312  1.1547005  0.24634591
#>  [7,]  0.42959613 -0.20852864 -0.8660254  0.85179945
#>  [8,]  0.42959613 -0.10106603 -0.8660254  0.52504674
#>  [9,]  0.42959613  0.11385920 -0.8660254  0.57309861
#> [10,]  0.90016715  0.01714285 -0.8660254  2.15304403
#> [11,]  1.10184044 -0.19778238 -0.8660254  2.07712208
#> [12,] -1.29134932  0.39326200  1.1547005 -1.50754727
#> [13,] -1.11656580  1.13475403  1.1547005 -1.29611905
#> [14,]  0.02624954 -0.49330457 -0.8660254  0.32322890
#> [15,]  0.02624954 -0.26225995 -0.8660254  0.24154072
#> [16,]  1.30351374 -1.27778166 -0.8660254  0.63076085
#> [17,] -1.10312091  0.59744096  1.1547005 -1.20001532
#> [18,] -0.76699875 -0.58464779 -0.8660254 -1.00300266
#> [19,] -0.47121125 -0.47718518  1.1547005 -1.60557308
#> [20,]  1.55896658 -1.76673655 -0.8660254 -0.01313418
#> [21,]  2.51355351 -1.71300524 -0.8660254  0.37128077
#> attr(,"center")
#>  [1]   0.4285714   2.8095238   6.1904762 227.3190476   3.6490476   3.7619048
#>  [7] 148.0476190  17.7880952   0.4285714   3.1836667
#> attr(,"scale")
#>  [1]   0.4948717   1.5922146   1.8418171 117.8966602   0.5494318   0.7499055
#>  [7]  74.3777208   1.8611124   0.4948717   1.0405423
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 33
#> 
#> $lambda.min
#> [1] 0.6124344
#> 
#> $null.dev
#> [1] 39.45048
#> 
#> $Bias
#> [1] 2.964364
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 10.02135