Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 33.213601 31.783381 29.024476 25.099014 21.857273 19.198577 17.036147
#>   [8] 15.295236 13.911514 12.829666 12.002175 11.388265 10.952986 10.666421
#>  [15] 10.502999 10.408055 10.104911  9.903267  9.880896  9.923131  9.988104
#>  [22] 10.003901  9.934749  9.824897  9.743233  9.685147  9.646698  9.624512
#>  [29]  9.615715  9.626083  9.689748 10.345916 10.644127 11.069451 10.640743
#>  [36] 10.583434 10.085287  9.474800  9.547945  9.571035  9.941346  9.957426
#>  [43] 15.580011 15.513089 16.129837 16.393985 16.540435 18.244772 18.282061
#>  [50] 18.272573 17.447157 17.183857 18.387483 18.379723 18.182339 18.171920
#>  [57] 18.165802 17.954172 17.738460 17.738947 17.739381 18.044465 21.970704
#>  [64] 21.971810 21.972884 21.973929 21.974948 21.975940 23.474954 23.475278
#>  [71] 22.060293 22.060489 22.060707 22.060947 22.061208 22.061489 22.061789
#>  [78] 22.062117 21.988402 21.793177 21.792881 21.792765 21.865375 21.865226
#>  [85] 21.865093 21.864978 21.868597 21.868521 21.868459 21.868411 21.868374
#>  [92] 21.868350 21.868338 21.868336 21.868321 21.867469 21.867522 21.867585
#>  [99] 21.867655 21.868037
#> 
#> $cvse
#>   [1] 11.056104 11.091405 10.498682  9.329991  8.341729  7.501058  6.782930
#>   [8]  6.169156  5.647377  5.209920  4.852560  4.573143  4.370122  4.241170
#>  [15]  4.182153  4.167945  4.015486  3.882470  3.841921  3.838460  3.834059
#>  [22]  3.832643  3.835414  3.777968  3.726184  3.679100  3.636015  3.596414
#>  [29]  3.559908  3.527167  3.502676  3.444443  3.411000  3.663078  3.704905
#>  [36]  3.708282  3.262787  3.303082  3.282917  3.264628  3.279407  3.260822
#>  [43]  5.656346  5.655829  5.720422  5.685623  5.626986  5.814881  5.808738
#>  [50]  5.805595  5.731015  5.745644  5.931024  5.928559  5.931378  5.932448
#>  [57]  5.930448  5.749850  5.678297  5.678456  5.678602  5.652441  7.601949
#>  [64]  7.602528  7.603094  7.603650  7.604194  7.604727  7.873408  7.873779
#>  [71]  7.871289  7.871664  7.872034  7.872398  7.872757  7.873112  7.873461
#>  [78]  7.873806  7.880573  7.814156  7.814380  7.814620  7.819827  7.820046
#>  [85]  7.820264  7.820479  7.820259  7.820469  7.820678  7.820884  7.821088
#>  [92]  7.821291  7.821491  7.821689  7.821888  7.822128  7.822317  7.822505
#>  [99]  7.822690  7.823027
#> 
#> $fold
#>  [1] 10  6  9  6  2  8 10  2  4  5  7  3  3  9  1  5  8  4  1  1  7
#> 
#> $lambda
#>   [1] 4.914274283 4.583068644 4.274185156 3.986119382 3.717468277 3.466923357
#>   [7] 3.233264326 3.015353132 2.812128423 2.622600379 2.445845891 2.281004064
#>  [13] 2.127272024 1.983901010 1.850192721 1.725495924 1.609203274 1.500748359
#>  [19] 1.399602943 1.305274389 1.217303263 1.135261098 1.058748299 0.987392207
#>  [25] 0.920845277 0.858783388 0.800904261 0.746925995 0.696585683 0.649638139
#>  [31] 0.605854703 0.565022123 0.526941523 0.491427428 0.458306864 0.427418516
#>  [37] 0.398611938 0.371746828 0.346692336 0.323326433 0.301535313 0.281212842
#>  [43] 0.262260038 0.244584589 0.228100406 0.212727202 0.198390101 0.185019272
#>  [49] 0.172549592 0.160920327 0.150074836 0.139960294 0.130527439 0.121730326
#>  [55] 0.113526110 0.105874830 0.098739221 0.092084528 0.085878339 0.080090426
#>  [61] 0.074692599 0.069658568 0.064963814 0.060585470 0.056502212 0.052694152
#>  [67] 0.049142743 0.045830686 0.042741852 0.039861194 0.037174683 0.034669234
#>  [73] 0.032332643 0.030153531 0.028121284 0.026226004 0.024458459 0.022810041
#>  [79] 0.021272720 0.019839010 0.018501927 0.017254959 0.016092033 0.015007484
#>  [85] 0.013996029 0.013052744 0.012173033 0.011352611 0.010587483 0.009873922
#>  [91] 0.009208453 0.008587834 0.008009043 0.007469260 0.006965857 0.006496381
#>  [97] 0.006058547 0.005650221 0.005269415 0.004914274
#> 
#> $fit
#> $beta
#>               4.9143     4.5831     4.2742    3.9861    3.7175   3.4669
#> (Intercept) 20.29048 21.8844112 23.3709202 24.757244 26.050133 27.25589
#> am           0.00000  0.0000000  0.0000000  0.000000  0.000000  0.00000
#> carb         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.00000
#> cyl          0.00000  0.0000000  0.0000000  0.000000  0.000000  0.00000
#> disp         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.00000
#> drat         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.00000
#> gear         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.00000
#> hp           0.00000  0.0000000  0.0000000  0.000000  0.000000  0.00000
#> qsec         0.00000  0.0000000  0.0000000  0.000000  0.000000  0.00000
#> vs           0.00000  0.0000000  0.0000000  0.000000  0.000000  0.00000
#> wt           0.00000 -0.5071458 -0.9801116 -1.421201 -1.832563 -2.21620
#>                3.2333    3.0154    2.8121    2.6226    2.4458    2.2810
#> (Intercept) 28.380376 29.429079 30.407103 31.319211 32.169846 32.963151
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.573981 -2.907649 -3.218829 -3.509037 -3.779685 -4.032093
#>                2.1273   1.9839    1.8502    1.7255    1.6092    1.5007
#> (Intercept) 33.702990 34.39297 35.036441 35.636547 36.057184 36.057184
#> am           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> wt          -4.267489 -4.48702 -4.691756 -4.882693 -5.016528 -5.016528
#>                1.3996    1.3053    1.2173      1.1353     1.0587     0.9874
#> (Intercept) 36.057184 36.057184 36.057184 35.23349821 34.0435751 32.9338393
#> am           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> carb         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> cyl          0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> disp         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> qsec         0.000000  0.000000  0.000000  0.04370398  0.1068402  0.1657217
#> vs           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> wt          -5.016528 -5.016528 -5.016528 -5.00506871 -4.9885148 -4.9730765
#>                 0.9208     0.8588     0.8009     0.7469     0.6966     0.6496
#> (Intercept) 31.8988959 30.9337043 30.0335633 29.1940889 28.4111922 27.6810601
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.2206349  0.2718471  0.3196077  0.3641494  0.4056892  0.4444293
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -4.9586786 -4.9452511 -4.9327286 -4.9210500 -4.9101586 -4.9000011
#>                 0.6059     0.5650    0.5269     0.4914     0.4583     0.4274
#> (Intercept) 27.0001365 26.3651048 25.772872 15.0486131 13.5728845 13.5706458
#> am           0.0000000  0.0000000  0.000000  2.4926351  2.9102471  2.9108321
#> carb         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> qsec         0.4805585  0.5142527  0.545676  0.9176802  0.9629734  0.9630459
#> vs           0.0000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> wt          -4.8905283 -4.8816939 -4.873455 -3.9343987 -3.7815364 -3.7813199
#>                0.3986     0.3717     0.3467      0.3233     0.3015     0.2812
#> (Intercept) 13.569224 13.5683206 13.5677469 14.14726732 15.3485931 14.9101469
#> am           2.911204  2.9114396  2.9115895  3.00387199  3.1981740  3.1269042
#> carb         0.000000  0.0000000  0.0000000 -0.07800601 -0.2406937 -0.5390506
#> cyl          0.000000  0.0000000  0.0000000  0.00000000  0.0000000  0.0000000
#> disp         0.000000  0.0000000  0.0000000  0.00000000  0.0000000  0.0000000
#> drat         0.000000  0.0000000  0.0000000  0.00000000  0.0000000  0.8990159
#> gear         0.000000  0.0000000  0.0000000  0.00000000  0.0000000  0.0000000
#> hp           0.000000  0.0000000  0.0000000  0.00000000  0.0000000  0.0000000
#> qsec         0.963092  0.9631213  0.9631399  0.92650233  0.8503343  0.6668328
#> vs           0.000000  0.0000000  0.0000000  0.00000000  0.0000000  0.0000000
#> wt          -3.781182 -3.7810950 -3.7810396 -3.69818654 -3.5247054 -3.0884453
#>                 0.2623     0.2446      0.2281     0.2127     0.1984      0.1850
#> (Intercept) 14.1740564 13.4868415 11.84965756 11.2985049 10.7723921 -6.80684063
#> am           3.0063108  2.8935196  1.96456490  1.9020709  1.8367635  2.05406177
#> carb        -0.5712977 -0.6014512 -1.09341757 -1.1015817 -1.1112180 -0.71269923
#> cyl          0.0000000  0.0000000  0.00000000  0.0000000  0.0000000  0.00000000
#> disp         0.0000000  0.0000000  0.00000000  0.0000000  0.0000000  0.02993528
#> drat         1.2316738  1.5425136  0.07912023  0.4021064  0.7107121  2.25696371
#> gear         0.0000000  0.0000000  2.19320293  2.1144322  2.0442168  2.12836512
#> hp           0.0000000  0.0000000  0.00000000  0.0000000  0.0000000  0.00000000
#> qsec         0.6354204  0.6060453  0.53254455  0.5100831  0.4876960  1.19646864
#> vs           0.0000000  0.0000000  0.00000000  0.0000000  0.0000000  0.00000000
#> wt          -3.0106804 -2.9380412 -2.37457798 -2.3310489 -2.2879475 -5.05398874
#>                  0.1725      0.1609      0.1501      0.1400      0.1305
#> (Intercept) -6.80894349 -6.81092719 -5.41600577 -5.41816473 -5.42012730
#> am           2.05424799  2.05442366  2.45318206  2.45330489  2.45341739
#> carb        -0.71256743 -0.71244310  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.02993973  0.02994392  0.05008880  0.05009314  0.05009706
#> drat         2.25706816  2.25716670  2.48405462  2.48415095  2.48423841
#> gear         2.12821701  2.12807729  2.00816606  2.00821567  2.00825882
#> hp           0.00000000  0.00000000 -0.02852796 -0.02852674 -0.02852561
#> qsec         1.19660753  1.19673854  1.08962711  1.08971948  1.08980388
#> vs           0.00000000  0.00000000  1.48222473  1.48228515  1.48233797
#> wt          -5.05450384 -5.05498975 -6.00932956 -6.00971936 -6.01007286
#>                  0.1217      0.1135      0.1059      0.0987      0.0921
#> (Intercept) -5.42190984 -5.42352754 -5.42499450 -5.42632375 -5.42752734
#> am           2.45352033  2.45361439  2.45370027  2.45377858  2.45384994
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.05010059  0.05010377  0.05010663  0.05010921  0.05011153
#> drat         2.48431773  2.48438963  2.48445475  2.48451369  2.48456699
#> gear         2.00829631  2.00832885  2.00835706  2.00838147  2.00840258
#> hp          -0.02852455 -0.02852356 -0.02852264 -0.02852179 -0.02852101
#> qsec         1.08988091  1.08995115  1.09001514  1.09007337  1.09012631
#> vs           1.48238410  1.48242436  1.48245946  1.48249002  1.48251662
#> wt          -6.01039317 -6.01068322 -6.01094566 -6.01118296 -6.01139739
#>                  0.0859        0.0801       0.0747       0.0697       0.0650
#> (Intercept) -5.42861636 -5.4307460642 -15.59110735 -15.58613754 -15.58139398
#> am           2.45391488  2.4539739501   3.04703217   3.04675382   3.04648885
#> carb         0.00000000  0.0000000000   0.00000000   0.00000000   0.00000000
#> cyl          0.00000000  0.0002511721   0.73845702   0.73815852   0.73788112
#> disp         0.05011361  0.0501121756   0.05073294   0.05073007   0.05072710
#> drat         2.48461517  2.4847068473   3.07727166   3.07698309   3.07670887
#> gear         2.00842079  2.0083802102   2.19821934   2.19803947   2.19786526
#> hp          -0.02852028 -0.0285203156  -0.03313375  -0.03313088  -0.03312821
#> qsec         1.09017441  1.0902410072   1.30014564   1.30004973   1.29995645
#> vs           1.48253973  1.4825440378   1.92308424   1.92277512   1.92247908
#> wt          -6.01159103 -6.0117337571  -6.26336038  -6.26313074  -6.26289516
#>                   0.0606       0.0565       0.0527       0.0491       0.0458
#> (Intercept) -15.57687590 -15.57258160 -15.56850846 -15.56465312 -15.56101154
#> am            3.04623716   3.04599858   3.04577289   3.04555984   3.04535915
#> carb          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> cyl           0.73762396   0.73738621   0.73716701   0.73696550   0.73678085
#> disp          0.05072406   0.05072097   0.05071783   0.05071469   0.05071154
#> drat          3.07644883   3.07620275   3.07597037   3.07575138   3.07554546
#> gear          2.19769696   2.19753475   2.19737877   2.19722913   2.19708587
#> hp           -0.03312574  -0.03312345  -0.03312133  -0.03311939  -0.03311760
#> qsec          1.29986597   1.29977843   1.29969393   1.29961258   1.29953442
#> vs            1.92219618   1.92192641   1.92166970   1.92142592   1.92119492
#> wt           -6.26265534  -6.26241283  -6.26216903  -6.26192521  -6.26168251
#>                   0.0427       0.0399       0.0372       0.0347       0.0323
#> (Intercept) -15.55757909 -15.55435063 -15.55132058 -15.54848303 -15.54583174
#> am            3.04517051   3.04499356   3.04482797   3.04467335   3.04452932
#> carb          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> cyl           0.73661219   0.73645871   0.73631958   0.73619400   0.73608119
#> disp          0.05070841   0.05070530   0.05070223   0.05069922   0.05069626
#> drat          3.07535224   3.07517133   3.07500235   3.07484486   3.07469845
#> gear          2.19694902   2.19681857   2.19669449   2.19657670   2.19646512
#> hp           -0.03311597  -0.03311448  -0.03311313  -0.03311191  -0.03311081
#> qsec          1.29945950   1.29938785   1.29931946   1.29925432   1.29919242
#> vs            1.92097646   1.92077030   1.92057617   1.92039374   1.92022269
#> wt           -6.26144195  -6.26120443  -6.26097075  -6.26074163  -6.26051765
#>                   0.0302       0.0281       0.0262       0.0245       0.0228
#> (Intercept) -15.54336023 -15.54106186 -15.53892983 -15.53695727 -15.53513725
#> am            3.04439546   3.04427139   3.04415669   3.04405095   3.04395375
#> carb          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> cyl           0.73598038   0.73589083   0.73581182   0.73574264   0.73568263
#> disp          0.05069337   0.05069055   0.05068782   0.05068516   0.05068260
#> drat          3.07456267   3.07443708   3.07432124   3.07421470   3.07411702
#> gear          2.19635965   2.19626015   2.19616649   2.19607851   2.19599605
#> hp           -0.03310983  -0.03310895  -0.03310818  -0.03310750  -0.03310691
#> qsec          1.29913370   1.29907813   1.29902564   1.29897617   1.29892964
#> vs            1.92006267   1.91991330   1.91977420   1.91964499   1.91952526
#> wt           -6.26029936  -6.26008719  -6.25988151  -6.25968263  -6.25949079
#>                   0.0213       0.0198       0.0185       0.0173       0.0161
#> (Intercept) -15.53346283 -15.53192710 -15.53052320 -15.52924434 -15.52808384
#> am            3.04386468   3.04378334   3.04370932   3.04364223   3.04358169
#> carb          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> cyl           0.73563115   0.73558756   0.73555127   0.73552172   0.73549836
#> disp          0.05068012   0.05067774   0.05067545   0.05067326   0.05067117
#> drat          3.07402776   3.07394648   3.07387276   3.07380617   3.07374630
#> gear          2.19591893   2.19584699   2.19578001   2.19571783   2.19566024
#> hp           -0.03310639  -0.03310596  -0.03310560  -0.03310530  -0.03310506
#> qsec          1.29888598   1.29884509   1.29880688   1.29877127   1.29873816
#> vs            1.91941462   1.91931266   1.91921899   1.91913319   1.91905488
#> wt           -6.25930616  -6.25912889  -6.25895905  -6.25879669  -6.25864180
#>                   0.0150       0.0140       0.0131       0.0122       0.0114
#> (Intercept) -15.52703515 -15.52609184 -15.52524766 -15.52449652 -15.52383252
#> am            3.04352730   3.04347871   3.04343555   3.04339748   3.04336415
#> carb          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> cyl           0.73548067   0.73546816   0.73546037   0.73545685   0.73545719
#> disp          0.05066917   0.05066727   0.05066546   0.05066375   0.05066213
#> drat          3.07369276   3.07364515   3.07360310   3.07356623   3.07353420
#> gear          2.19560705   2.19555805   2.19551306   2.19547188   2.19543432
#> hp           -0.03310487  -0.03310474  -0.03310465  -0.03310460  -0.03310459
#> qsec          1.29870745   1.29867903   1.29865281   1.29862870   1.29860658
#> vs            1.91898366   1.91891915   1.91886097   1.91880874   1.91876212
#> wt           -6.25849436  -6.25835430  -6.25822152  -6.25809591  -6.25797733
#>                   0.0106       0.0099       0.0092       0.0086       0.0080
#> (Intercept) -15.52324993 -15.52274325 -15.52230716 -15.52193655 -15.52162652
#> am            3.04333525   3.04331046   3.04328948   3.04327202   3.04325782
#> carb          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> cyl           0.73546101   0.73546792   0.73547758   0.73548967   0.73550389
#> disp          0.05066060   0.05065916   0.05065780   0.05065653   0.05065534
#> drat          3.07350667   3.07348331   3.07346381   3.07344786   3.07343519
#> gear          2.19540018   2.19536928   2.19534143   2.19531646   2.19529419
#> hp           -0.03310462  -0.03310467  -0.03310475  -0.03310486  -0.03310499
#> qsec          1.29858636   1.29856795   1.29855124   1.29853614   1.29852256
#> vs            1.91872075   1.91868430   1.91865243   1.91862484   1.91860121
#> wt           -6.25786561  -6.25776060  -6.25766210  -6.25756991  -6.25748382
#>                   0.0075       0.0070       0.0065       0.0061       0.0057
#> (Intercept) -15.52137241 -15.52116973 -15.52101424 -15.52090188 -15.52082881
#> am            3.04324659   3.04323811   3.04323212   3.04322841   3.04322676
#> carb          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> cyl           0.73551996   0.73553761   0.73555660   0.73557671   0.73559773
#> disp          0.05065423   0.05065320   0.05065224   0.05065134   0.05065052
#> drat          3.07342552   3.07341859   3.07341416   3.07341199   3.07341187
#> gear          2.19527446   2.19525709   2.19524193   2.19522883   2.19521763
#> hp           -0.03310513  -0.03310529  -0.03310546  -0.03310565  -0.03310584
#> qsec          1.29851041   1.29849960   1.29849005   1.29848167   1.29847439
#> vs            1.91858126   1.91856471   1.91855129   1.91854075   1.91853284
#> wt           -6.25740361  -6.25732908  -6.25725998  -6.25719610  -6.25713720
#>                   0.0053        0.0049
#> (Intercept) -15.52079140 -1.552316e+01
#> am            3.04322697  3.043229e+00
#> carb          0.00000000 -2.198556e-04
#> cyl           0.73561947  7.357727e-01
#> disp          0.05064976  5.064863e-02
#> drat          3.07341359  3.073633e+00
#> gear          2.19520821  2.195397e+00
#> hp           -0.03310604 -3.310274e-02
#> qsec          1.29846812  1.298473e+00
#> vs            1.91852734  1.918523e+00
#> wt           -6.25708306 -6.256999e+00
#> 
#> $iter
#>   [1]   0   3   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   4   3   3   3   3   3   3   3   3   3   3   3  81  13   1
#>  [37]   1   1   1  33  38  70  23  23  45  21  21 124   1   1 140   1   1   1
#>  [55]   1   1   1   1   1   1  97   1   1   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.914274283 4.583068644 4.274185156 3.986119382 3.717468277 3.466923357
#>   [7] 3.233264326 3.015353132 2.812128423 2.622600379 2.445845891 2.281004064
#>  [13] 2.127272024 1.983901010 1.850192721 1.725495924 1.609203274 1.500748359
#>  [19] 1.399602943 1.305274389 1.217303263 1.135261098 1.058748299 0.987392207
#>  [25] 0.920845277 0.858783388 0.800904261 0.746925995 0.696585683 0.649638139
#>  [31] 0.605854703 0.565022123 0.526941523 0.491427428 0.458306864 0.427418516
#>  [37] 0.398611938 0.371746828 0.346692336 0.323326433 0.301535313 0.281212842
#>  [43] 0.262260038 0.244584589 0.228100406 0.212727202 0.198390101 0.185019272
#>  [49] 0.172549592 0.160920327 0.150074836 0.139960294 0.130527439 0.121730326
#>  [55] 0.113526110 0.105874830 0.098739221 0.092084528 0.085878339 0.080090426
#>  [61] 0.074692599 0.069658568 0.064963814 0.060585470 0.056502212 0.052694152
#>  [67] 0.049142743 0.045830686 0.042741852 0.039861194 0.037174683 0.034669234
#>  [73] 0.032332643 0.030153531 0.028121284 0.026226004 0.024458459 0.022810041
#>  [79] 0.021272720 0.019839010 0.018501927 0.017254959 0.016092033 0.015007484
#>  [85] 0.013996029 0.013052744 0.012173033 0.011352611 0.010587483 0.009873922
#>  [91] 0.009208453 0.008587834 0.008009043 0.007469260 0.006965857 0.006496381
#>  [97] 0.006058547 0.005650221 0.005269415 0.004914274
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 33
#> 
#> $loss
#>   [1] 656.03810 558.68026 477.22595 409.38654 353.18610 306.91976 269.11741
#>   [8] 238.51232 214.01381 194.68360 179.71525 168.41633 160.19298 154.53653
#>  [15] 151.01180 149.24714 148.88617 148.88617 148.88617 148.88617 148.88617
#>  [22] 144.94379 139.73487 135.39501 131.79821 128.83569 126.41368 124.45134
#>  [29] 122.87908 121.63703 120.67373 119.94499 119.41290 102.07792 101.72680
#>  [36] 101.72680 101.72680 101.72680 101.72680  99.91076  97.22675  91.73417
#>  [43]  91.04842  90.83761  85.81658  84.64166  83.91769  68.76594  68.76593
#>  [50]  68.76592  59.57835  59.57834  59.57833  59.57832  59.57832  59.57831
#>  [57]  59.57831  59.57831  59.57830  59.57688  57.48425  57.48423  57.48422
#>  [64]  57.48421  57.48420  57.48419  57.48419  57.48418  57.48417  57.48416
#>  [71]  57.48416  57.48415  57.48415  57.48414  57.48414  57.48413  57.48413
#>  [78]  57.48413  57.48412  57.48412  57.48412  57.48412  57.48412  57.48411
#>  [85]  57.48411  57.48411  57.48411  57.48411  57.48411  57.48411  57.48411
#>  [92]  57.48410  57.48410  57.48410  57.48410  57.48410  57.48410  57.48410
#>  [99]  57.48410  57.48402
#> 
#> $linear.predictors
#>      4.9143   4.5831   4.2742   3.9861   3.7175   3.4669   3.2333   3.0154
#> 1  20.29048 20.55569 20.80303 21.03370 21.24882 21.44944 21.63655 21.81104
#> 2  20.29048 20.70783 21.09706 21.46006 21.79859 22.11430 22.40874 22.68333
#> 3  20.29048 20.25394 20.21986 20.18808 20.15844 20.13080 20.10503 20.08099
#> 4  20.29048 20.13983 19.99934 19.86831 19.74612 19.63216 19.52588 19.42677
#> 5  20.29048 20.12969 19.97973 19.83989 19.70947 19.58784 19.47440 19.36861
#> 6  20.29048 20.28690 20.28357 20.28046 20.27756 20.27486 20.27234 20.26998
#> 7  20.29048 20.13983 19.99934 19.86831 19.74612 19.63216 19.52588 19.42677
#> 8  20.29048 20.13983 19.99934 19.86831 19.74612 19.63216 19.52588 19.42677
#> 9  20.29048 19.96740 19.66610 19.38510 19.12305 18.87865 18.65073 18.43817
#> 10 20.29048 19.13365 18.05479 17.04865 16.11031 15.23522 14.41910 13.65799
#> 11 20.29048 19.17372 18.13222 17.16092 16.25509 15.41030 14.62245 13.88769
#> 12 20.29048 20.95380 21.57242 22.14934 22.68738 23.18916 23.65712 24.09354
#> 13 20.29048 20.63430 20.95495 21.25398 21.53287 21.79295 22.03551 22.26172
#> 14 20.29048 20.09926 19.92093 19.75462 19.59951 19.45486 19.31996 19.19415
#> 15 20.29048 19.93697 19.60729 19.29983 19.01309 18.74568 18.49629 18.26371
#> 16 20.29048 20.90308 21.47440 22.00722 22.50412 22.96754 23.39972 23.80278
#> 17 20.29048 20.79912 21.27348 21.71587 22.12845 22.51322 22.87206 23.20671
#> 18 20.29048 21.11710 21.88801 22.60697 23.27747 23.90278 24.48594 25.02981
#> 19 20.29048 20.47962 20.65601 20.82052 20.97393 21.11701 21.25045 21.37489
#> 20 20.29048 20.07390 19.87192 19.68356 19.50788 19.34405 19.19126 19.04877
#> 21 20.29048 20.47455 20.64621 20.80630 20.95561 21.09485 21.22471 21.34581
#>      2.8121   2.6226   2.4458   2.2810   2.1273   1.9839    1.8502    1.7255
#> 1  21.97377 22.12554 22.26707 22.39907 22.52217 22.63697 22.744041 22.843892
#> 2  22.93942 23.17825 23.40098 23.60870 23.80242 23.98308 24.151568 24.308700
#> 3  20.05857 20.03766 20.01816 19.99997 19.98301 19.96720 19.952446 19.938690
#> 4  19.33433 19.24813 19.16773 19.09275 19.02283 18.95762 18.896801 18.840084
#> 5  19.26995 19.17794 19.09214 19.01211 18.93748 18.86788 18.802966 18.742430
#> 6  20.26779 20.26575 20.26384 20.26206 20.26040 20.25885 20.257410 20.256065
#> 7  19.33433 19.24813 19.16773 19.09275 19.02283 18.95762 18.896801 18.840084
#> 8  19.33433 19.24813 19.16773 19.09275 19.02283 18.95762 18.896801 18.840084
#> 9  18.23993 18.05505 17.88264 17.72184 17.57188 17.43203 17.301604 17.179968
#> 10 12.94817 12.28620 11.66883 11.09308 10.55613 10.05537  9.588358  9.152821
#> 11 13.20246 12.56341 11.96743 11.41162 10.89326 10.40984  9.959007  9.538554
#> 12 24.50055 24.88013 25.23412 25.56426 25.87215 26.15928 26.427069 26.676806
#> 13 22.47269 22.66944 22.85292 23.02404 23.18363 23.33246 23.471263 23.600709
#> 14 19.07682 18.96740 18.86535 18.77019 18.68143 18.59866 18.521461 18.449468
#> 15 18.04680 17.84451 17.65586 17.47992 17.31583 17.16281 17.020099 16.887007
#> 16 24.17867 24.52923 24.85616 25.16105 25.44540 25.71058 25.957894 26.188537
#> 17 23.51881 23.80987 24.08132 24.33447 24.57056 24.79074 24.996084 25.187585
#> 18 25.53701 26.01004 26.45118 26.86260 27.24628 27.60411 27.937815 28.249033
#> 19 21.49095 21.59918 21.70012 21.79425 21.88205 21.96392 22.040278 22.111488
#> 20 18.91588 18.79195 18.67637 18.56858 18.46806 18.37430 18.286873 18.205334
#> 21 21.45876 21.56409 21.66232 21.75393 21.83937 21.91905 21.993360 22.062661
#>       1.6092    1.5007    1.3996    1.3053    1.2173    1.1353    1.0587
#> 1  22.913881 22.913881 22.913881 22.913881 22.913881 22.839586 22.732256
#> 2  24.418840 24.418840 24.418840 24.418840 24.418840 24.435070 24.458517
#> 3  19.929047 19.929047 19.929047 19.929047 19.929047 19.991808 20.082473
#> 4  18.800329 18.800329 18.800329 18.800329 18.800329 18.759904 18.701504
#> 5  18.699998 18.699998 18.699998 18.699998 18.699998 18.799655 18.943623
#> 6  20.255122 20.255122 20.255122 20.255122 20.255122 20.468353 20.776394
#> 7  18.800329 18.800329 18.800329 18.800329 18.800329 18.815845 18.838260
#> 8  18.800329 18.800329 18.800329 18.800329 18.800329 18.842067 18.902364
#> 9  17.094709 17.094709 17.094709 17.094709 17.094709 17.101010 17.110113
#> 10  8.847538  8.847538  8.847538  8.847538  8.847538  8.864810  8.889763
#> 11  9.243844  9.243844  9.243844  9.243844  9.243844  9.242729  9.241120
#> 12 26.851855 26.851855 26.851855 26.851855 26.851855 26.918906 27.015770
#> 13 23.691443 23.691443 23.691443 23.691443 23.691443 23.770520 23.884758
#> 14 18.399006 18.399006 18.399006 18.399006 18.399006 18.352943 18.286397
#> 15 16.793718 16.793718 16.793718 16.793718 16.793718 16.687513 16.534086
#> 16 26.350203 26.350203 26.350203 26.350203 26.350203 26.374695 26.410079
#> 17 25.321815 25.321815 25.321815 25.321815 25.321815 25.252508 25.152385
#> 18 28.467177 28.467177 28.467177 28.467177 28.467177 28.399427 28.301552
#> 19 22.161402 22.161402 22.161402 22.161402 22.161402 22.046870 21.881412
#> 20 18.148180 18.148180 18.148180 18.148180 18.148180 18.003481 17.794444
#> 21 22.111237 22.111237 22.111237 22.111237 22.111237 22.132301 22.162732
#>       0.9874    0.9208    0.8588    0.8009    0.7469    0.6966    0.6496
#> 1  22.632159 22.538808 22.451749 22.370557 22.294837 22.224221 22.158364
#> 2  24.480384 24.500777 24.519795 24.537532 24.554074 24.569500 24.583887
#> 3  20.167029 20.245886 20.319429 20.388015 20.451978 20.511631 20.567263
#> 4  18.647040 18.596247 18.548877 18.504700 18.463500 18.425077 18.389243
#> 5  19.077888 19.203105 19.319883 19.428790 19.530357 19.625079 19.713417
#> 6  21.063676 21.331597 21.581461 21.814485 22.031803 22.234475 22.423488
#> 7  18.859164 18.878660 18.896842 18.913798 18.929611 18.944359 18.958113
#> 8  18.958597 19.011041 19.059950 19.105563 19.148101 19.187773 19.224771
#> 9  17.118602 17.126519 17.133902 17.140788 17.147210 17.153198 17.158784
#> 10  8.913034  8.934737  8.954977  8.973853  8.991456  9.007874  9.023185
#> 11  9.239618  9.238219  9.236913  9.235695  9.234560  9.233501  9.232513
#> 12 27.106107 27.190355 27.268925 27.342200 27.410536 27.474266 27.533702
#> 13 23.991298 24.090657 24.183320 24.269737 24.350331 24.425492 24.495588
#> 14 18.224336 18.166458 18.112480 18.062141 18.015194 17.971411 17.930579
#> 15 16.390998 16.257554 16.133103 16.017040 15.908800 15.807854 15.713712
#> 16 26.443077 26.473852 26.502553 26.529319 26.554281 26.577561 26.599272
#> 17 25.059009 24.971926 24.890713 24.814973 24.744337 24.678463 24.617027
#> 18 28.210272 28.125145 28.045755 27.971715 27.902666 27.838270 27.778214
#> 19 21.727105 21.583197 21.448988 21.323824 21.207097 21.098236 20.996712
#> 20 17.599494 17.417683 17.248125 17.089995 16.942522 16.804988 16.676724
#> 21 22.191111 22.217578 22.242261 22.265281 22.286749 22.306771 22.325442
#>       0.6059    0.5650    0.5269   0.4914   0.4583   0.4274   0.3986   0.3717
#> 1  22.096945 22.039666 21.986247 22.33814 22.42605 22.42616 22.42622 22.42627
#> 2  24.597305 24.609818 24.621487 25.49147 25.63090 25.63110 25.63123 25.63131
#> 3  20.619145 20.667531 20.712656 20.23922 20.13545 20.13532 20.13523 20.13518
#> 4  18.355825 18.324659 18.295593 17.13320 16.95421 16.95395 16.95378 16.95368
#> 5  19.795802 19.872633 19.944287 19.99109 19.96009 19.96007 19.96005 19.96004
#> 6  22.599762 22.764156 22.917470 23.67013 23.71314 23.71324 23.71331 23.71335
#> 7  18.970940 18.982902 18.994058 18.30783 18.18681 18.18665 18.18654 18.18647
#> 8  19.259275 19.291454 19.321464 18.85844 18.76460 18.76447 18.76440 18.76435
#> 9  17.163993 17.168850 17.173381 16.69483 16.61220 16.61208 16.61201 16.61196
#> 10  9.037464  9.050780  9.063199 10.06150 10.22202 10.22225 10.22239 10.22248
#> 11  9.231592  9.230733  9.229932 10.00524 10.13557 10.13575 10.13587 10.13594
#> 12 27.589131 27.640825 27.689035 28.58346 28.70718 28.70737 28.70749 28.70756
#> 13 24.560960 24.621926 24.678783 23.71310 23.52049 23.52024 23.52008 23.51998
#> 14 17.892499 17.856985 17.823865 16.68079 16.50724 16.50698 16.50682 16.50672
#> 15 15.625914 15.544034 15.467673 14.08197 13.89120 13.89092 13.89073 13.89061
#> 16 26.619520 26.638403 26.656013 27.27234 27.36606 27.36619 27.36628 27.36633
#> 17 24.559733 24.506300 24.456468 24.44689 24.47230 24.47232 24.47233 24.47234
#> 18 27.722206 27.669972 27.621259 27.09730 27.03592 27.03582 27.03575 27.03571
#> 19 20.902030 20.813729 20.731380 20.86701 20.93436 20.93443 20.93448 20.93451
#> 20 16.557105 16.445547 16.341508 16.89358 17.04246 17.04264 17.04275 17.04282
#> 21 22.342856 22.359096 22.374241 23.67247 23.88177 23.88206 23.88225 23.88237
#>      0.3467   0.3233   0.3015   0.2812   0.2623   0.2446   0.2281   0.2127
#> 1  22.42630 22.40009 22.34577 22.27135 22.26974 22.26820 21.06622 21.10881
#> 2  25.63136 25.73555 25.95348 26.20378 26.19141 26.17984 26.19986 26.18945
#> 3  20.13514 20.19080 20.30647 20.17394 20.06955 19.97205 20.29794 20.20040
#> 4  16.95361 17.03856 17.21491 17.38919 17.36935 17.35089 17.38702 17.36808
#> 5  19.96004 20.00741 20.10618 19.64972 19.43342 19.23134 20.10623 19.89848
#> 6  23.71338 23.55887 23.23704 22.89806 22.92711 22.95420 23.46113 23.46743
#> 7  18.18643 18.06847 17.82195 17.85688 17.98848 18.11146 18.13597 18.24188
#> 8  18.76431 18.62438 18.33215 18.25698 18.36973 18.47509 18.45549 18.54793
#> 9  16.61194 16.61115 16.60914 16.38164 16.29860 16.22102 16.00181 15.94165
#> 10 10.22254 10.28655 10.42077 10.58223 10.57715 10.57237 10.90319 10.88787
#> 11 10.13599 10.20811 10.35909 10.76625 10.84411 10.91684 10.89596 10.96047
#> 12 28.70761 28.72436 28.75989 28.89452 28.92700 28.95731 28.06779 28.12679
#> 13 23.51991 23.49254 23.43469 23.42776 23.45339 23.47739 22.43148 22.48874
#> 14 16.50666 16.60373 16.80538 16.69147 16.55283 16.42336 17.08632 16.94826
#> 15 13.89054 13.91161 13.95460 14.52354 14.71382 14.89169 13.43885 13.64448
#> 16 27.36637 27.42804 27.55709 27.79298 27.81808 27.84151 27.28671 27.32731
#> 17 24.47235 24.55360 24.72309 25.46843 25.66275 25.84434 26.75580 26.88085
#> 18 27.03569 27.05766 27.10315 26.94490 26.86463 26.78964 28.29895 28.17904
#> 19 20.93453 20.79991 20.51935 19.83810 19.72067 19.61088 20.18301 20.06815
#> 20 17.04287 16.85150 16.45290 15.61717 15.49912 15.38870 15.61089 15.50891
#> 21 23.88245 23.94711 24.08292 24.47112 24.54908 24.62188 24.02938 24.11503
#>      0.1984   0.1850   0.1725   0.1609   0.1501   0.1400   0.1305   0.1217
#> 1  21.14598 20.95411 20.95418 20.95425 21.82461 21.82447 21.82435 21.82424
#> 2  26.17903 25.51133 25.51123 25.51113 25.20848 25.20847 25.20846 25.20846
#> 3  20.10788 20.55108 20.55118 20.55127 21.38880 21.38894 21.38907 21.38918
#> 4  17.35140 19.01717 19.01741 19.01763 19.34620 19.34650 19.34677 19.34702
#> 5  19.70031 18.53601 18.53591 18.53581 18.46123 18.46117 18.46112 18.46107
#> 6  23.47402 24.82247 24.82264 24.82279 24.20180 24.20189 24.20197 24.20205
#> 7  18.34468 17.22993 17.22969 17.22945 17.99040 17.99011 17.98984 17.98959
#> 8  18.63729 17.94781 17.94765 17.94750 18.64418 18.64394 18.64372 18.64352
#> 9  15.88336 15.05754 15.05749 15.05744 13.81202 13.81191 13.81182 13.81173
#> 10 10.87322 11.17682 11.17683 11.17685 11.79055 11.79061 11.79068 11.79073
#> 11 11.02236 11.01789 11.01783 11.01777 10.97107 10.97109 10.97110 10.97111
#> 12 28.18078 29.23642 29.23662 29.23681 29.39823 29.39837 29.39850 29.39862
#> 13 22.54247 22.29480 22.29482 22.29483 21.52061 21.52055 21.52049 21.52044
#> 14 16.81803 16.29588 16.29583 16.29578 16.34270 16.34270 16.34271 16.34271
#> 15 13.84080 14.65355 14.65363 14.65372 14.13107 14.13117 14.13125 14.13133
#> 16 27.36479 27.45507 27.45510 27.45513 27.72708 27.72711 27.72714 27.72716
#> 17 27.00457 27.22870 27.22852 27.22835 26.84881 26.84879 26.84878 26.84876
#> 18 28.06759 28.39288 28.39287 28.39286 28.78749 28.78764 28.78778 28.78791
#> 19 19.95738 18.66939 18.66927 18.66916 18.58414 18.58390 18.58368 18.58348
#> 20 15.40881 16.61333 16.61362 16.61390 15.84667 15.84690 15.84710 15.84729
#> 21 24.19527 23.43780 23.43767 23.43755 23.27386 23.27377 23.27369 23.27361
#>      0.1135   0.1059   0.0987   0.0921   0.0859   0.0801   0.0747   0.0697
#> 1  21.82414 21.82405 21.82397 21.82390 21.82383 21.82395 22.14385 22.14379
#> 2  25.20845 25.20844 25.20844 25.20843 25.20843 25.20831 25.03564 25.03571
#> 3  21.38929 21.38938 21.38946 21.38954 21.38961 21.38961 21.41789 21.41778
#> 4  19.34724 19.34744 19.34762 19.34778 19.34793 19.34810 19.65258 19.65234
#> 5  18.46102 18.46098 18.46094 18.46091 18.46088 18.46092 18.40423 18.40428
#> 6  24.20212 24.20219 24.20225 24.20230 24.20235 24.20230 24.18083 24.18084
#> 7  17.98937 17.98917 17.98898 17.98881 17.98866 17.98872 18.29260 18.29252
#> 8  18.64334 18.64318 18.64303 18.64289 18.64277 18.64287 19.07268 19.07255
#> 9  13.81165 13.81158 13.81152 13.81147 13.81142 13.81172 14.11362 14.11364
#> 10 11.79078 11.79082 11.79086 11.79090 11.79093 11.79071 11.55254 11.55256
#> 11 10.97112 10.97114 10.97114 10.97115 10.97116 10.97097 10.72340 10.72347
#> 12 29.39872 29.39882 29.39891 29.39899 29.39906 29.39918 29.94485 29.94460
#> 13 21.52039 21.52035 21.52032 21.52028 21.52026 21.52016 20.72215 20.72259
#> 14 16.34271 16.34271 16.34271 16.34271 16.34271 16.34288 16.45392 16.45387
#> 15 14.13140 14.13147 14.13153 14.13158 14.13163 14.13170 14.01213 14.01220
#> 16 27.72718 27.72720 27.72722 27.72724 27.72726 27.72727 27.95521 27.95509
#> 17 26.84874 26.84872 26.84871 26.84869 26.84868 26.84845 26.43001 26.43013
#> 18 28.78802 28.78812 28.78821 28.78829 28.78836 28.78821 28.50183 28.50181
#> 19 18.58330 18.58313 18.58298 18.58285 18.58273 18.58272 18.37811 18.37830
#> 20 15.84746 15.84762 15.84776 15.84789 15.84801 15.84809 16.04096 16.04086
#> 21 23.27355 23.27349 23.27344 23.27339 23.27335 23.27316 23.07097 23.07108
#>      0.0650   0.0606   0.0565   0.0527   0.0491   0.0458   0.0427   0.0399
#> 1  22.14374 22.14370 22.14367 22.14365 22.14364 22.14364 22.14365 22.14366
#> 2  25.03576 25.03582 25.03587 25.03592 25.03596 25.03600 25.03604 25.03607
#> 3  21.41767 21.41755 21.41744 21.41732 21.41720 21.41709 21.41697 21.41686
#> 4  19.65210 19.65186 19.65162 19.65139 19.65115 19.65092 19.65070 19.65048
#> 5  18.40432 18.40437 18.40441 18.40446 18.40450 18.40454 18.40459 18.40463
#> 6  24.18085 24.18085 24.18085 24.18085 24.18085 24.18084 24.18083 24.18082
#> 7  18.29247 18.29243 18.29241 18.29240 18.29241 18.29243 18.29246 18.29250
#> 8  19.07244 19.07235 19.07228 19.07222 19.07218 19.07215 19.07214 19.07213
#> 9  14.11368 14.11372 14.11377 14.11382 14.11388 14.11395 14.11402 14.11409
#> 10 11.55256 11.55256 11.55255 11.55253 11.55251 11.55249 11.55246 11.55243
#> 11 10.72352 10.72358 10.72362 10.72366 10.72370 10.72373 10.72376 10.72378
#> 12 29.94435 29.94412 29.94390 29.94369 29.94349 29.94330 29.94313 29.94296
#> 13 20.72301 20.72340 20.72378 20.72413 20.72446 20.72477 20.72506 20.72533
#> 14 16.45383 16.45379 16.45375 16.45371 16.45368 16.45365 16.45362 16.45359
#> 15 14.01226 14.01231 14.01236 14.01240 14.01244 14.01247 14.01249 14.01251
#> 16 27.95499 27.95488 27.95479 27.95470 27.95461 27.95453 27.95445 27.95438
#> 17 26.43025 26.43035 26.43045 26.43054 26.43062 26.43069 26.43075 26.43081
#> 18 28.50178 28.50174 28.50169 28.50163 28.50156 28.50149 28.50141 28.50133
#> 19 18.37849 18.37869 18.37888 18.37907 18.37926 18.37945 18.37964 18.37982
#> 20 16.04075 16.04065 16.04054 16.04044 16.04033 16.04023 16.04012 16.04002
#> 21 23.07118 23.07128 23.07137 23.07147 23.07155 23.07164 23.07172 23.07180
#>      0.0372   0.0347   0.0323   0.0302   0.0281   0.0262   0.0245   0.0228
#> 1  22.14368 22.14370 22.14373 22.14377 22.14380 22.14384 22.14388 22.14392
#> 2  25.03610 25.03613 25.03616 25.03618 25.03620 25.03622 25.03623 25.03625
#> 3  21.41675 21.41663 21.41653 21.41642 21.41632 21.41622 21.41612 21.41603
#> 4  19.65026 19.65005 19.64985 19.64965 19.64946 19.64928 19.64911 19.64894
#> 5  18.40467 18.40471 18.40474 18.40478 18.40482 18.40485 18.40488 18.40491
#> 6  24.18081 24.18079 24.18078 24.18076 24.18074 24.18072 24.18071 24.18069
#> 7  18.29255 18.29260 18.29266 18.29273 18.29280 18.29287 18.29295 18.29302
#> 8  19.07214 19.07215 19.07218 19.07221 19.07224 19.07228 19.07233 19.07238
#> 9  14.11417 14.11425 14.11433 14.11441 14.11449 14.11457 14.11465 14.11473
#> 10 11.55240 11.55236 11.55233 11.55229 11.55225 11.55220 11.55216 11.55212
#> 11 10.72380 10.72381 10.72382 10.72383 10.72384 10.72384 10.72384 10.72384
#> 12 29.94280 29.94265 29.94251 29.94238 29.94226 29.94215 29.94205 29.94195
#> 13 20.72558 20.72581 20.72603 20.72623 20.72642 20.72659 20.72674 20.72689
#> 14 16.45357 16.45355 16.45352 16.45350 16.45349 16.45347 16.45345 16.45344
#> 15 14.01253 14.01254 14.01255 14.01255 14.01256 14.01255 14.01255 14.01255
#> 16 27.95432 27.95426 27.95420 27.95415 27.95410 27.95406 27.95402 27.95398
#> 17 26.43086 26.43090 26.43094 26.43098 26.43101 26.43103 26.43105 26.43107
#> 18 28.50124 28.50115 28.50107 28.50097 28.50088 28.50079 28.50070 28.50061
#> 19 18.37999 18.38017 18.38033 18.38049 18.38065 18.38080 18.38095 18.38108
#> 20 16.03992 16.03982 16.03972 16.03962 16.03953 16.03944 16.03935 16.03927
#> 21 23.07188 23.07196 23.07203 23.07209 23.07216 23.07222 23.07227 23.07233
#>      0.0213   0.0198   0.0185   0.0173   0.0161   0.0150   0.0140   0.0131
#> 1  22.14397 22.14402 22.14406 22.14411 22.14416 22.14421 22.14425 22.14430
#> 2  25.03626 25.03627 25.03628 25.03629 25.03629 25.03630 25.03630 25.03631
#> 3  21.41594 21.41586 21.41577 21.41570 21.41562 21.41555 21.41548 21.41542
#> 4  19.64878 19.64863 19.64848 19.64835 19.64821 19.64809 19.64797 19.64786
#> 5  18.40494 18.40497 18.40500 18.40502 18.40505 18.40507 18.40509 18.40511
#> 6  24.18067 24.18065 24.18063 24.18061 24.18059 24.18057 24.18056 24.18054
#> 7  18.29311 18.29319 18.29327 18.29335 18.29343 18.29351 18.29360 18.29368
#> 8  19.07244 19.07249 19.07255 19.07261 19.07268 19.07274 19.07280 19.07287
#> 9  14.11480 14.11488 14.11496 14.11503 14.11510 14.11517 14.11524 14.11531
#> 10 11.55208 11.55204 11.55199 11.55195 11.55191 11.55187 11.55183 11.55179
#> 11 10.72384 10.72383 10.72383 10.72382 10.72381 10.72380 10.72379 10.72378
#> 12 29.94186 29.94178 29.94170 29.94163 29.94157 29.94151 29.94146 29.94141
#> 13 20.72702 20.72713 20.72724 20.72734 20.72742 20.72750 20.72756 20.72762
#> 14 16.45343 16.45342 16.45341 16.45340 16.45339 16.45338 16.45338 16.45337
#> 15 14.01254 14.01253 14.01252 14.01251 14.01250 14.01248 14.01247 14.01245
#> 16 27.95395 27.95391 27.95389 27.95386 27.95384 27.95382 27.95380 27.95379
#> 17 26.43108 26.43109 26.43110 26.43110 26.43110 26.43110 26.43110 26.43110
#> 18 28.50052 28.50043 28.50034 28.50026 28.50018 28.50010 28.50002 28.49994
#> 19 18.38122 18.38134 18.38146 18.38158 18.38169 18.38179 18.38189 18.38198
#> 20 16.03919 16.03911 16.03904 16.03897 16.03890 16.03884 16.03877 16.03872
#> 21 23.07238 23.07242 23.07247 23.07251 23.07255 23.07259 23.07262 23.07265
#>      0.0122   0.0114   0.0106   0.0099   0.0092   0.0086   0.0080   0.0075
#> 1  22.14435 22.14439 22.14444 22.14448 22.14453 22.14457 22.14461 22.14465
#> 2  25.03631 25.03631 25.03631 25.03631 25.03631 25.03631 25.03631 25.03630
#> 3  21.41536 21.41530 21.41524 21.41519 21.41515 21.41510 21.41506 21.41502
#> 4  19.64776 19.64766 19.64757 19.64749 19.64741 19.64734 19.64727 19.64721
#> 5  18.40513 18.40515 18.40517 18.40518 18.40520 18.40521 18.40522 18.40523
#> 6  24.18052 24.18050 24.18049 24.18047 24.18045 24.18044 24.18042 24.18041
#> 7  18.29375 18.29383 18.29390 18.29398 18.29405 18.29412 18.29418 18.29424
#> 8  19.07293 19.07299 19.07306 19.07312 19.07318 19.07324 19.07329 19.07335
#> 9  14.11537 14.11543 14.11549 14.11554 14.11560 14.11565 14.11570 14.11574
#> 10 11.55176 11.55172 11.55169 11.55165 11.55162 11.55159 11.55156 11.55153
#> 11 10.72377 10.72376 10.72375 10.72374 10.72372 10.72371 10.72370 10.72369
#> 12 29.94137 29.94134 29.94130 29.94127 29.94125 29.94123 29.94121 29.94119
#> 13 20.72767 20.72772 20.72775 20.72778 20.72781 20.72783 20.72784 20.72786
#> 14 16.45337 16.45336 16.45336 16.45336 16.45335 16.45335 16.45335 16.45335
#> 15 14.01244 14.01242 14.01241 14.01239 14.01238 14.01236 14.01234 14.01233
#> 16 27.95377 27.95376 27.95375 27.95374 27.95374 27.95373 27.95373 27.95373
#> 17 26.43109 26.43108 26.43107 26.43106 26.43105 26.43104 26.43103 26.43102
#> 18 28.49987 28.49980 28.49973 28.49967 28.49960 28.49954 28.49949 28.49943
#> 19 18.38207 18.38215 18.38223 18.38230 18.38236 18.38243 18.38248 18.38254
#> 20 16.03866 16.03861 16.03856 16.03851 16.03847 16.03843 16.03839 16.03836
#> 21 23.07268 23.07271 23.07273 23.07275 23.07278 23.07279 23.07281 23.07283
#>      0.0070   0.0065   0.0061   0.0057   0.0053   0.0049
#> 1  22.14469 22.14472 22.14476 22.14479 22.14482 22.14459
#> 2  25.03630 25.03630 25.03629 25.03629 25.03629 25.03638
#> 3  21.41498 21.41495 21.41492 21.41489 21.41486 21.41487
#> 4  19.64715 19.64709 19.64705 19.64700 19.64696 19.64718
#> 5  18.40524 18.40525 18.40526 18.40527 18.40528 18.40526
#> 6  24.18040 24.18038 24.18037 24.18036 24.18035 24.18029
#> 7  18.29431 18.29436 18.29442 18.29447 18.29452 18.29440
#> 8  19.07341 19.07346 19.07351 19.07356 19.07361 19.07348
#> 9  14.11579 14.11583 14.11587 14.11590 14.11594 14.11606
#> 10 11.55150 11.55148 11.55145 11.55143 11.55140 11.55134
#> 11 10.72367 10.72366 10.72365 10.72364 10.72363 10.72367
#> 12 29.94118 29.94117 29.94116 29.94116 29.94115 29.94124
#> 13 20.72786 20.72787 20.72787 20.72786 20.72786 20.72774
#> 14 16.45335 16.45335 16.45335 16.45335 16.45335 16.45346
#> 15 14.01231 14.01230 14.01228 14.01227 14.01225 14.01242
#> 16 27.95372 27.95372 27.95372 27.95372 27.95373 27.95378
#> 17 26.43101 26.43099 26.43098 26.43097 26.43095 26.43110
#> 18 28.49938 28.49933 28.49929 28.49925 28.49921 28.49925
#> 19 18.38258 18.38263 18.38267 18.38271 18.38274 18.38244
#> 20 16.03833 16.03830 16.03827 16.03824 16.03822 16.03817
#> 21 23.07284 23.07285 23.07286 23.07287 23.07288 23.07289
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 22.8 21.4 18.7 18.1 22.8 19.2 17.8 15.2 10.4 14.7 33.9 21.5 15.5 13.3
#> [16] 27.3 26.0 30.4 19.7 15.0 21.4
#> 
#> $X
#>               am       carb         cyl        disp        drat       gear
#>  [1,]  1.1547005  0.6556285  0.05643326 -0.46060418  0.61515882  0.3175003
#>  [2,]  1.1547005 -0.9965553 -1.12866530 -0.90833408  0.50996065  0.3175003
#>  [3,] -0.8660254 -0.9965553  0.05643326  0.38319447 -1.11009116 -1.0160010
#>  [4,] -0.8660254 -0.4458274  1.24153183  1.26143389 -0.96281372 -1.0160010
#>  [5,] -0.8660254 -0.9965553  0.05643326  0.09905819 -1.78335945 -1.0160010
#>  [6,] -0.8660254 -0.4458274 -1.12866530 -0.62591984  0.65723809  0.3175003
#>  [7,] -0.8660254  0.6556285  0.05643326 -0.39516674  0.65723809  0.3175003
#>  [8,] -0.8660254  0.6556285  0.05643326 -0.39516674  0.65723809  0.3175003
#>  [9,] -0.8660254  0.1049006  1.24153183  0.53645586 -1.13113080 -1.0160010
#> [10,] -0.8660254  0.6556285  1.24153183  2.12245292 -1.27840823 -1.0160010
#> [11,] -0.8660254  0.6556285  1.24153183  1.95024911 -0.79449665 -1.0160010
#> [12,]  1.1547005 -0.9965553 -1.12866530 -1.22605010  1.28842711  0.3175003
#> [13,] -0.8660254 -0.9965553 -1.12866530 -0.80415078  0.19436614 -1.0160010
#> [14,] -0.8660254 -0.4458274  1.24153183  0.89980589 -1.78335945 -1.0160010
#> [15,] -0.8660254  0.6556285  1.24153183  1.17533198  0.25748504 -1.0160010
#> [16,]  1.1547005 -0.9965553 -1.12866530 -1.15802960  0.99387223  0.3175003
#> [17,]  1.1547005 -0.4458274 -1.12866530 -0.80242874  1.73025942  1.6510017
#> [18,]  1.1547005 -0.4458274 -1.12866530 -1.01940554  0.34164358  1.6510017
#> [19,]  1.1547005  1.7570843  0.05643326 -0.58975704  0.02604907  1.6510017
#> [20,]  1.1547005  2.8585401  1.24153183  0.75343266 -0.14226800  1.6510017
#> [21,]  1.1547005 -0.4458274 -1.12866530 -0.79640161  1.05699113  0.3175003
#>               hp        qsec         vs           wt
#>  [1,] -0.5010985 -0.83544464 -1.0488088 -0.533833666
#>  [2,] -0.7584973  0.31386448  0.9534626 -0.840075890
#>  [3,] -0.5010985  0.75755126  0.9534626  0.073546743
#>  [4,]  0.4830734 -0.53608970 -1.0488088  0.303228411
#>  [5,] -0.5768041  1.17450992  0.9534626  0.323644559
#>  [6,] -0.7282151  2.60713711  0.9534626  0.007194262
#>  [7,] -0.3042641  0.14815015  0.9534626  0.303228411
#>  [8,] -0.3042641  0.46888758  0.9534626  0.303228411
#>  [9,]  0.5587789 -0.01221857 -1.0488088  0.650302931
#> [10,]  1.0887177 -0.10843980 -1.0488088  2.328510315
#> [11,]  1.3158343 -0.32226475 -1.0488088  2.247866530
#> [12,] -1.1824483  1.00344996  0.9534626 -1.335167484
#> [13,] -0.6979329  1.06225182  0.9534626 -0.692058815
#> [14,]  0.1045457 -0.61627406 -1.0488088  0.384893004
#> [15,]  1.5429509 -1.39673514 -1.0488088  0.711551376
#> [16,] -1.1673072  0.46888758  0.9534626 -1.233086743
#> [17,] -0.7887796 -0.70714967 -1.0488088 -1.023821224
#> [18,] -0.4556752 -0.60023719  0.9534626 -1.663867471
#> [19,]  0.4830734 -1.34862453 -1.0488088 -0.380712555
#> [20,]  2.9056505 -1.82973067 -1.0488088  0.435933375
#> [21,] -0.5162396  0.30851886  0.9534626 -0.370504480
#> attr(,"center")
#>  [1]   0.4285714   2.8095238   5.9047619 213.4952381   3.6076190   3.7619048
#>  [7] 143.0952381  18.0228571   0.5238095   3.1429524
#> attr(,"scale")
#>  [1]   0.4948717   1.8157787   1.6876233 116.1414510   0.4752934   0.7499055
#>  [7]  66.0453702   1.8706891   0.4994328   0.9796167
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 38
#> 
#> $lambda.min
#> [1] 0.3717468
#> 
#> $null.dev
#> [1] 31.23991
#> 
#> $Bias
#> [1] 5.95551
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#>  7.15948