Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 31.99643 30.03331 28.11901 25.50923 23.36249 21.61018 20.19320 19.06064
#>   [9] 18.16873 17.47985 16.96170 16.58659 16.33078 16.17400 16.09892 16.13000
#>  [17] 15.99781 15.75241 15.54439 15.42659 15.31001 15.16725 14.85277 14.52232
#>  [25] 14.25136 13.93660 13.60567 13.37878 13.40339 13.48668 13.70111 14.01672
#>  [33] 14.48553 15.73196 16.55548 16.58919 17.66576 17.63853 17.44884 17.13465
#>  [41] 17.70402 17.80123 18.18492 23.05821 23.03941 25.83794 28.16234 28.98769
#>  [49] 27.97788 31.31551 39.01811 39.71281 43.03805 42.64421 43.34830 43.35111
#>  [57] 43.35380 43.36921 43.25521 42.36834 42.03989 41.74706 41.86994 42.08626
#>  [65] 42.08832 42.08137 42.08324 42.10698 42.10867 42.11030 42.11022 41.86502
#>  [73] 41.86709 41.86908 41.87098 41.83996 40.76315 40.76377 40.76435 40.80592
#>  [81] 40.80649 40.60111 40.60158 40.49233 40.49252 40.49271 40.49287 40.49303
#>  [89] 40.49318 40.49308 40.45218 40.45215 40.45212 40.45209 40.45206 40.45202
#>  [97] 40.45198 40.45195 40.45191 40.45187
#> 
#> $cvse
#>   [1] 11.226969 11.157787 11.102490 10.275368  9.527247  8.847207  8.227873
#>   [8]  7.664581  7.154686  6.696985  6.291223  5.937642  5.636556  5.387950
#>  [15]  5.191131  5.040162  4.859756  4.629591  4.452529  4.359640  4.277737
#>  [22]  4.183445  4.082655  3.982312  3.893820  3.802363  3.710974  3.634149
#>  [29]  3.553018  3.487978  3.456707  3.480242  3.835556  4.372209  4.386359
#>  [36]  4.369571  4.999130  4.976720  5.040993  4.946026  5.745506  5.798360
#>  [43]  5.773647  6.869937  6.867455  7.882480  8.248483  8.199549  8.221002
#>  [50]  9.416955 13.602223 13.534510 14.853571 14.901315 15.104552 15.105926
#>  [57] 15.107234 15.083009 15.086099 14.822598 14.840165 14.879710 14.865940
#>  [64] 14.846581 14.847515 14.849635 14.850508 14.849108 14.849925 14.850713
#>  [71] 14.851775 14.823304 14.824459 14.825566 14.826626 14.831729 14.321139
#>  [78] 14.321657 14.322150 14.320252 14.320698 14.346959 14.347367 14.356001
#>  [85] 14.356386 14.356753 14.357101 14.357432 14.357746 14.358063 14.361492
#>  [92] 14.361771 14.362036 14.362288 14.362527 14.362754 14.362969 14.363173
#>  [99] 14.363367 14.363550
#> 
#> $fold
#>  [1]  3  2  9  5  8  1  4  5 10  6  1 10  3  7  8  9  6  7  2  4  1
#> 
#> $lambda
#>   [1] 4.602487741 4.292295471 4.003009122 3.733219705 3.481613192 3.246964115
#>   [7] 3.028129601 2.824043801 2.633712700 2.456209279 2.290668994 2.136285571
#>  [13] 1.992307073 1.858032244 1.732807090 1.616021691 1.507107238 1.405533254
#>  [19] 1.310805017 1.222461146 1.140071356 1.063234362 0.991575925 0.924747026
#>  [25] 0.862422172 0.804297804 0.750090824 0.699537213 0.652390746 0.608421793
#>  [31] 0.567416200 0.529174248 0.493509674 0.460248774 0.429229547 0.400300912
#>  [37] 0.373321971 0.348161319 0.324696411 0.302812960 0.282404380 0.263371270
#>  [43] 0.245620928 0.229066899 0.213628557 0.199230707 0.185803224 0.173280709
#>  [49] 0.161602169 0.150710724 0.140553325 0.131080502 0.122246115 0.114007136
#>  [55] 0.106323436 0.099157592 0.092474703 0.086242217 0.080429780 0.075009082
#>  [61] 0.069953721 0.065239075 0.060842179 0.056741620 0.052917425 0.049350967
#>  [67] 0.046024877 0.042922955 0.040030091 0.037332197 0.034816132 0.032469641
#>  [73] 0.030281296 0.028240438 0.026337127 0.024562093 0.022906690 0.021362856
#>  [79] 0.019923071 0.018580322 0.017328071 0.016160217 0.015071072 0.014055333
#>  [85] 0.013108050 0.012224611 0.011400714 0.010632344 0.009915759 0.009247470
#>  [91] 0.008624222 0.008042978 0.007500908 0.006995372 0.006523907 0.006084218
#>  [97] 0.005674162 0.005291742 0.004935097 0.004602488
#> 
#> $fit
#> $beta
#>             4.6025     4.2923     4.0030    3.7332    3.4816    3.2470
#> (Intercept)   20.2 21.7766736 23.2470847 24.618395 25.897284 27.089979
#> am             0.0  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> carb           0.0  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> cyl            0.0  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> disp           0.0  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> drat           0.0  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> gear           0.0  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> hp             0.0  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> qsec           0.0  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> vs             0.0  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> wt             0.0 -0.4901431 -0.9472522 -1.373554 -1.771124 -2.141899
#>                3.0281    2.8240    2.6337    2.4562    2.2907    2.1363
#> (Intercept) 28.202291 29.239637 30.207070 31.109300 31.950724 32.735438
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.487685 -2.810167 -3.110914 -3.391392 -3.652967 -3.896912
#>                1.9923    1.8580    1.7328    1.6160    1.5071    1.4055
#> (Intercept) 33.467265 34.149769 34.786275 35.379883 35.795964 35.795964
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -4.124416 -4.336587 -4.534459 -4.718995 -4.848343 -4.848343
#>                  1.3108     1.2225     1.1401     1.0632     0.9916     0.9247
#> (Intercept) 34.35209446 32.9207965 31.5859613 30.3410897 29.1801181 28.0973922
#> am           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.07678344  0.1528984  0.2238835  0.2900845  0.3518237  0.4094019
#> vs           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -4.82864076 -4.8091105 -4.7908965 -4.7739101 -4.7580685 -4.7432945
#>                 0.8624     0.8043     0.7501     0.6995     0.6524     0.6084
#> (Intercept) 27.0876384 26.1459386 25.2677063 24.4486638 23.6848221 22.9724607
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.4630996  0.5131782  0.5598817  0.6034375  0.6440578  0.6819404
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -4.7295163 -4.7166667 -4.7046831 -4.6935072 -4.6830845 -4.6733642
#>                 0.5674     0.5292    0.4935       0.4602     0.4292     0.4003
#> (Intercept) 22.3081102 21.6885346 21.110716 20.814104135 17.9710685 15.3267381
#> am           0.0000000  0.0000000  0.000000  0.001312916  0.8328476  1.6063088
#> carb         0.0000000  0.0000000  0.000000  0.000000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.000000  0.000000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.000000  0.000000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.000000  0.000000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.000000  0.000000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.000000  0.000000000  0.0000000  0.0000000
#> qsec         0.7172699  0.7502183  0.780946  0.796616037  0.8822442  0.9618841
#> vs           0.0000000  0.0000000  0.000000  0.000000000  0.0000000  0.0000000
#> wt          -4.6642991 -4.6558449 -4.647960 -4.643490067 -4.3367380 -4.0514112
#>                   0.3733     0.3482     0.3247     0.3028     0.2824     0.2634
#> (Intercept) 12.939145381 13.7254575 14.4579837 15.1418975 15.4591521 11.9066116
#> am           2.315784735  2.5122569  2.6950185  2.8659308  2.9439433  1.9401764
#> carb        -0.006220962 -0.2455864 -0.4684147 -0.6766240 -0.7724702 -1.0944997
#> cyl          0.000000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.000000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.000000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.000000000  0.0000000  0.0000000  0.0000000  0.0000000  1.3185532
#> hp           0.000000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         1.033063505  0.9814393  0.9333637  0.8884606  0.8677136  0.7769146
#> vs           0.000000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -3.786058977 -3.5738868 -3.3764244 -3.1918639 -3.1071514 -2.6219466
#>                  0.2456      0.2291      0.2136      0.1992      0.1858
#> (Intercept) 25.86690188 25.86800968 25.86872650 25.86918427 25.86947239
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb        -1.19666859 -1.19663856 -1.19662393 -1.19661791 -1.19661647
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         2.02841620  2.02819875  2.02805833  2.02796885  2.02791266
#> hp          -0.03086739 -0.03086398 -0.03086162 -0.03086001 -0.03085892
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -1.75002450 -1.75029600 -1.75047493 -1.75059145 -1.75066637
#>                  0.1733      0.1616     0.1507      0.1406      0.1311
#> (Intercept) 25.86965077 25.86975909 25.8698233 20.64363252 10.63395335
#> am           0.00000000  0.00000000  0.0000000  1.40332027  2.14803804
#> carb        -1.19661724 -1.19661893 -1.1966208 -1.12951972  0.00000000
#> cyl          0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.0000000  0.00000000  0.02945499
#> drat         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> gear         2.02787797  2.02785698  2.0278446  1.53986967  1.25349490
#> hp          -0.03085819 -0.03085771 -0.0308574 -0.02377058 -0.04322201
#> qsec         0.00000000  0.00000000  0.0000000  0.29962643  1.02329245
#> vs           0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> wt          -1.75071389 -1.75074357 -1.7507618 -1.77824676 -4.64452463
#>                  0.1222      0.1140      0.1063      0.0992      0.0925
#> (Intercept) 10.63302934 18.52291717 18.51843762 18.51432228 18.51056157
#> am           2.14824528  1.82723374  1.82752883  1.82782413  1.82811688
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000 -0.64205484 -0.64180804 -0.64159636 -0.64141717
#> disp         0.02945067  0.03261797  0.03262141  0.03262493  0.03262849
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         1.25331904  0.89382425  0.89401724  0.89417839  0.89431037
#> hp          -0.04321386 -0.03885171 -0.03885622 -0.03886014 -0.03886351
#> qsec         1.02334396  0.80125242  0.80141706  0.80157445  0.80172409
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -4.64439819 -4.61371558 -4.61401419 -4.61431508 -4.61461516
#>                  0.0862      0.0804      0.0750      0.0700      0.0652
#> (Intercept) 18.50714385 21.59189126 21.58737670 21.58312925 21.57914818
#> am           1.82840472  1.97368742  1.97378944  1.97390248  1.97402451
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl         -0.64126788 -0.73410468 -0.73384626 -0.73361399 -0.73340649
#> disp         0.03263205  0.03026208  0.03026485  0.03026777  0.03027080
#> drat         0.00000000 -0.96313785 -0.96279904 -0.96248172 -0.96218567
#> gear         0.89441578  1.12689286  1.12700989  1.12711019  1.12719496
#> hp          -0.03886637 -0.03493680 -0.03494247 -0.03494769 -0.03495249
#> qsec         0.80186561  0.79054547  0.79065971  0.79077183  0.79088127
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -4.61491178 -4.53324644 -4.53344836 -4.53366004 -4.53387887
#>                  0.0608      0.0567      0.0529      0.0494      0.0460
#> (Intercept) 21.57543097 21.57197355 22.62719764 22.62957509 22.63186112
#> am           1.97415365  1.97428818  2.03430888  2.03438584  2.03445942
#> carb         0.00000000  0.00000000  0.38766541  0.38792826  0.38818029
#> cyl         -0.73322241 -0.73306036 -0.93966954 -0.93995257 -0.94022428
#> disp         0.03027390  0.03027705  0.03578079  0.03578477  0.03578858
#> drat        -0.96191054 -0.96165589 -1.20918029 -1.20948368 -1.20977510
#> gear         1.12726538  1.12732259  0.89575960  0.89550251  0.89525584
#> hp          -0.03495687 -0.03496086 -0.03595714 -0.03595479 -0.03595252
#> qsec         0.79098759  0.79109038  0.89962953  0.89968575  0.89973949
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -4.53410252 -4.53432885 -5.20529594 -5.20578889 -5.20626122
#>                  0.0429      0.0400      0.0373      0.0348      0.0325
#> (Intercept) 22.63405894 22.63617165 22.63820226 22.64015372 22.64202884
#> am           2.03452978  2.03459704  2.03466134  2.03472281  2.03478157
#> carb         0.38842193  0.38865358  0.38887563  0.38908846  0.38929244
#> cyl         -0.94048509 -0.94073541 -0.94097563 -0.94120615 -0.94142732
#> disp         0.03579223  0.03579572  0.03579906  0.03580226  0.03580532
#> drat        -1.21005497 -1.21032372 -1.21058176 -1.21082949 -1.21106730
#> gear         0.89501918  0.89479216  0.89457441  0.89436556  0.89416528
#> hp          -0.03595035 -0.03594825 -0.03594623 -0.03594429 -0.03594242
#> qsec         0.89979086  0.89983995  0.89988686  0.89993169  0.89997453
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -5.20671375 -5.20714728 -5.20756255 -5.20796031 -5.20834127
#>                  0.0303      0.0282      0.0263      0.0246      0.0229
#> (Intercept) 22.64383039 22.64556102 22.64722332 22.64881978 22.65035283
#> am           2.03483773  2.03489140  2.03494269  2.03499170  2.03503854
#> carb         0.38948790  0.38967520  0.38985465  0.39002658  0.39019127
#> cyl         -0.94163950 -0.94184303 -0.94203825 -0.94222548 -0.94240503
#> disp         0.03580825  0.03581106  0.03581374  0.03581630  0.03581876
#> drat        -1.21129555 -1.21151460 -1.21172480 -1.21192650 -1.21212000
#> gear         0.89397324  0.89378910  0.89361258  0.89344336  0.89328116
#> hp          -0.03594062 -0.03593889 -0.03593723 -0.03593563 -0.03593410
#> qsec         0.90001546  0.90005457  0.90009193  0.90012762  0.90016171
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -5.20870609 -5.20905543 -5.20938993 -5.20971018 -5.21001677
#>                  0.0214      0.0199      0.0186      0.0173      0.0162
#> (Intercept) 22.65182481 22.65323799 22.65459454 22.65589660 22.65714621
#> am           2.03508329  2.03512604  2.03516688  2.03520590  2.03524317
#> carb         0.39034902  0.39050011  0.39064481  0.39078337  0.39091605
#> cyl         -0.94257719 -0.94274224 -0.94290048 -0.94305216 -0.94319755
#> disp         0.03582111  0.03582335  0.03582550  0.03582756  0.03582952
#> drat        -1.21230562 -1.21248367 -1.21265443 -1.21281819 -1.21297522
#> gear         0.89312570  0.89297672  0.89283397  0.89269718  0.89256614
#> hp          -0.03593262 -0.03593120 -0.03592984 -0.03592853 -0.03592727
#> qsec         0.90019428  0.90022538  0.90025508  0.90028345  0.90031054
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -5.21031026 -5.21059118 -5.21086005 -5.21111736 -5.21136360
#>                  0.0151     0.0141      0.0131      0.0122      0.0114
#> (Intercept) 22.65834535 22.6594959 22.66059979 22.66165872 22.66267445
#> am           2.03527876  2.0353128  2.03534523  2.03537624  2.03540585
#> carb         0.39104309  0.3911647  0.39128114  0.39139258  0.39149925
#> cyl         -0.94333689 -0.9434704 -0.94359836 -0.94372095 -0.94383840
#> disp         0.03583140  0.0358332  0.03583492  0.03583656  0.03583813
#> drat        -1.21312578 -1.2132701 -1.21340849 -1.21354112 -1.21366824
#> gear         0.89244060  0.8923203  0.89220516  0.89209485  0.89198921
#> hp          -0.03592606 -0.0359249 -0.03592379 -0.03592272 -0.03592169
#> qsec         0.90033640  0.9003611  0.90038467  0.90040718  0.90042867
#> vs           0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> wt          -5.21159922 -5.2118247 -5.21204034 -5.21224668 -5.21244406
#>                  0.0106      0.0099      0.0092      0.0086      0.0080
#> (Intercept) 22.66364863 22.66458286 22.66547870 22.66633764 22.66716111
#> am           2.03543412  2.03546111  2.03548688  2.03551148  2.03553497
#> carb         0.39160133  0.39169903  0.39179252  0.39188197  0.39196755
#> cyl         -0.94395091 -0.94405868 -0.94416190 -0.94426075 -0.94435542
#> disp         0.03583963  0.03584107  0.03584244  0.03584375  0.03584500
#> drat        -1.21379005 -1.21390679 -1.21401863 -1.21412579 -1.21422845
#> gear         0.89188805  0.89179119  0.89169846  0.89160969  0.89152471
#> hp          -0.03592070 -0.03591975 -0.03591885 -0.03591797 -0.03591714
#> qsec         0.90044918  0.90046875  0.90048744  0.90050527  0.90052229
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt          -5.21263285 -5.21281342 -5.21298612 -5.21315126 -5.21330918
#>                  0.0075       0.0070      0.0065      0.0061      0.0057
#> (Intercept) 22.66795050 22.668176565 22.67723657 22.67476244 22.67236042
#> am           2.03555738  2.035537263  2.04271498  2.04271173  2.04271342
#> carb         0.39204943  0.392342628  0.39248311  0.39239179  0.39231074
#> cyl         -0.94444607 -0.944557641 -0.93206176 -0.93178037 -0.93151210
#> disp         0.03584620  0.035848399  0.03584181  0.03584289  0.03584402
#> drat        -1.21432678 -1.214502481 -1.20138968 -1.20085522 -1.20036219
#> gear         0.89144337  0.891299098  0.88767730  0.88775149  0.88782651
#> hp          -0.03591633 -0.035915011 -0.03612195 -0.03612912 -0.03613594
#> qsec         0.90053852  0.900533863  0.89204372  0.89198168  0.89192792
#> vs           0.00000000  0.001373788  0.06270535  0.06313370  0.06354148
#> wt          -5.21346018 -5.213563147 -5.20275680 -5.20260565 -5.20247795
#>                  0.0053      0.0049      0.0046
#> (Intercept) 22.67003520 22.66778974 22.66562595
#> am           2.04272105  2.04273503  2.04275530
#> carb         0.39223951  0.39217761  0.39212452
#> cyl         -0.93125686 -0.93101455 -0.93078505
#> disp         0.03584521  0.03584645  0.03584775
#> drat        -1.19990632 -1.19948380 -1.19909138
#> gear         0.88789996  0.88797009  0.88803570
#> hp          -0.03614243 -0.03614860 -0.03615446
#> qsec         0.89188160  0.89184201  0.89180851
#> vs           0.06393039  0.06430165  0.06465618
#> wt          -5.20237209 -5.20228658 -5.20222000
#> 
#> $iter
#>   [1]   0   3   3   3   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   4   3   3   3   3   3   3   3   3   3   3   3   3   3   3   3  40  39
#>  [37]  38  35  34  34  13  34  39   1   1   1   1   1   1   1  86 110   1  75
#>  [55]   1   1   1   1  90   1   1   1   1   1 127   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   2  24   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.602487741 4.292295471 4.003009122 3.733219705 3.481613192 3.246964115
#>   [7] 3.028129601 2.824043801 2.633712700 2.456209279 2.290668994 2.136285571
#>  [13] 1.992307073 1.858032244 1.732807090 1.616021691 1.507107238 1.405533254
#>  [19] 1.310805017 1.222461146 1.140071356 1.063234362 0.991575925 0.924747026
#>  [25] 0.862422172 0.804297804 0.750090824 0.699537213 0.652390746 0.608421793
#>  [31] 0.567416200 0.529174248 0.493509674 0.460248774 0.429229547 0.400300912
#>  [37] 0.373321971 0.348161319 0.324696411 0.302812960 0.282404380 0.263371270
#>  [43] 0.245620928 0.229066899 0.213628557 0.199230707 0.185803224 0.173280709
#>  [49] 0.161602169 0.150710724 0.140553325 0.131080502 0.122246115 0.114007136
#>  [55] 0.106323436 0.099157592 0.092474703 0.086242217 0.080429780 0.075009082
#>  [61] 0.069953721 0.065239075 0.060842179 0.056741620 0.052917425 0.049350967
#>  [67] 0.046024877 0.042922955 0.040030091 0.037332197 0.034816132 0.032469641
#>  [73] 0.030281296 0.028240438 0.026337127 0.024562093 0.022906690 0.021362856
#>  [79] 0.019923071 0.018580322 0.017328071 0.016160217 0.015071072 0.014055333
#>  [85] 0.013108050 0.012224611 0.011400714 0.010632344 0.009915759 0.009247470
#>  [91] 0.008624222 0.008042978 0.007500908 0.006995372 0.006523907 0.006084218
#>  [97] 0.005674162 0.005291742 0.004935097 0.004602488
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 33
#> 
#> $loss
#>   [1] 623.06000 537.66402 466.21759 406.71326 357.41788 316.83605 283.67828
#>   [8] 256.83348 235.34498 218.38978 205.26051 195.34982 188.13684 183.17536
#>  [15] 180.08371 178.53586 178.21924 178.21924 170.52457 163.66607 157.95994
#>  [22] 153.23862 149.35756 146.19214 143.63497 141.59362 139.98860 138.75158
#>  [29] 137.82392 137.15534 136.70276 136.42937 136.30373 136.27537 129.87175
#>  [36] 126.43791 125.31528 121.73322 119.58918 118.62325 118.51548 112.97761
#>  [43] 108.39127 108.39127 108.39127 108.39126 108.39126 108.39126 108.39126
#>  [50] 108.39126 105.38133  96.48547  96.48546  93.65929  93.65928  93.65927
#>  [57]  93.65926  93.65925  92.92885  92.92884  92.92883  92.92882  92.92881
#>  [64]  92.92880  92.41006  92.41004  92.41003  92.41001  92.41000  92.40998
#>  [71]  92.40997  92.40996  92.40995  92.40994  92.40993  92.40992  92.40992
#>  [78]  92.40991  92.40990  92.40990  92.40989  92.40989  92.40988  92.40988
#>  [85]  92.40988  92.40987  92.40987  92.40987  92.40987  92.40986  92.40986
#>  [92]  92.40986  92.40986  92.40965  92.40495  92.40494  92.40493  92.40492
#>  [99]  92.40491  92.40490
#> 
#> $linear.predictors
#>    4.6025   4.2923   4.0030   3.7332   3.4816   3.2470   3.0281   2.8240
#> 1    20.2 20.49250 20.76528 21.01968 21.25694 21.47820 21.68456 21.87700
#> 2    20.2 20.36751 20.52373 20.66943 20.80530 20.93202 21.05020 21.16041
#> 3    20.2 20.63954 21.04946 21.43175 21.78828 22.12077 22.43086 22.72005
#> 4    20.2 20.20086 20.20167 20.20242 20.20312 20.20377 20.20438 20.20495
#> 5    20.2 20.09058 19.98854 19.89337 19.80462 19.72185 19.64465 19.57266
#> 6    20.2 20.21312 20.22535 20.23676 20.24740 20.25732 20.26658 20.27521
#> 7    20.2 20.23272 20.26324 20.29170 20.31824 20.34300 20.36608 20.38761
#> 8    20.2 20.09058 19.98854 19.89337 19.80462 19.72185 19.64465 19.57266
#> 9    20.2 19.94844 19.71383 19.49504 19.29099 19.10070 18.92323 18.75772
#> 10   20.2 19.92393 19.66647 19.42636 19.20244 18.99360 18.79884 18.61721
#> 11   20.2 19.11814 18.10919 17.16824 16.29071 15.47232 14.70909 13.99729
#> 12   20.2 19.15686 18.18402 17.27675 16.43063 15.64153 14.90561 14.21930
#> 13   20.2 20.87726 21.50888 22.09792 22.64727 23.15959 23.63739 24.08298
#> 14   20.2 20.56847 20.91211 21.23259 21.53146 21.81020 22.07015 22.31258
#> 15   20.2 20.09303 19.99327 19.90024 19.81347 19.73256 19.65709 19.58671
#> 16   20.2 19.89452 19.60964 19.34395 19.09617 18.86509 18.64958 18.44860
#> 17   20.2 19.89207 19.60490 19.33708 19.08731 18.85438 18.63714 18.43455
#> 18   20.2 20.72777 21.21997 21.67899 22.10708 22.50632 22.87864 23.22588
#> 19   20.2 21.03509 21.81389 22.54021 23.21757 23.84929 24.43842 24.98786
#> 20   20.2 20.22292 20.24430 20.26423 20.28282 20.30016 20.31633 20.33141
#> 21   20.2 20.41408 20.61372 20.79992 20.97356 21.13550 21.28653 21.42737
#>      2.6337   2.4562   2.2907   2.1363   1.9923   1.8580   1.7328    1.6160
#> 1  22.05647 22.22385 22.37995 22.52553 22.66129 22.78791 22.90599 23.016116
#> 2  21.26319 21.35905 21.44844 21.53182 21.60957 21.68208 21.74971 21.812773
#> 3  22.98975 23.24127 23.47584 23.69460 23.89862 24.08889 24.26633 24.431815
#> 4  20.20548 20.20598 20.20644 20.20687 20.20727 20.20764 20.20799 20.208314
#> 5  19.50553 19.44291 19.38452 19.33006 19.27927 19.23191 19.18774 19.146541
#> 6  20.28325 20.29076 20.29776 20.30429 20.31038 20.31606 20.32135 20.326289
#> 7  20.40769 20.42642 20.44388 20.46017 20.47535 20.48952 20.50273 20.515049
#> 8  19.50553 19.44291 19.38452 19.33006 19.27927 19.23191 19.18774 19.146541
#> 9  18.60336 18.45941 18.32516 18.19996 18.08319 17.97430 17.87274 17.778032
#> 10 18.44781 18.28984 18.14251 18.00511 17.87697 17.75747 17.64602 17.542082
#> 11 13.33347 12.71439 12.13703 11.59859 11.09643 10.62812 10.19137  9.784055
#> 12 13.57923 12.98231 12.42562 11.90644 11.42226 10.97071 10.54959 10.156856
#> 13 24.49854 24.88610 25.24753 25.58460 25.89896 26.19213 26.46554 26.720527
#> 14 22.53867 22.74952 22.94616 23.12955 23.30058 23.46008 23.60883 23.747561
#> 15 19.52108 19.45987 19.40278 19.34955 19.29990 19.25359 19.21041 19.170136
#> 16 18.26116 18.08636 17.92333 17.77130 17.62951 17.49727 17.37395 17.258943
#> 17 18.24561 18.06940 17.90507 17.75181 17.60888 17.47559 17.35128 17.235348
#> 18 23.54971 23.85172 24.13338 24.39605 24.64101 24.86947 25.08253 25.281234
#> 19 25.50026 25.97812 26.42379 26.83941 27.22702 27.58851 27.92564 28.240044
#> 20 20.34547 20.35859 20.37082 20.38223 20.39287 20.40279 20.41204 20.420669
#> 21 21.55873 21.68123 21.79548 21.90202 22.00139 22.09406 22.18048 22.261077
#>       1.5071    1.4055    1.3108   1.2225    1.1401    1.0632    0.9916
#> 1  23.093306 23.093306 22.964911 22.83763 22.718935 22.608236 22.504997
#> 2  21.856979 21.856979 21.776606 21.69693 21.622631 21.553336 21.488711
#> 3  24.547809 24.547809 24.578588 24.60910 24.637553 24.664090 24.688839
#> 4  20.208542 20.208542 20.320685 20.43185 20.535524 20.632211 20.722381
#> 5  19.117665 19.117665 19.048424 18.97979 18.915774 18.856077 18.800402
#> 6  20.329751 20.329751 20.484399 20.63770 20.780671 20.914006 21.038354
#> 7  20.523685 20.523685 20.900217 21.27347 21.621569 21.946207 22.248965
#> 8  19.117665 19.117665 19.192777 19.26724 19.336675 19.401435 19.461831
#> 9  17.711646 17.711646 17.692653 17.67383 17.656267 17.639892 17.624620
#> 10 17.469229 17.469229 17.481934 17.49453 17.506275 17.517230 17.527446
#> 11  9.498554  9.498554  9.529828  9.56083  9.589743  9.616707  9.641853
#> 12  9.881573  9.881573  9.880577  9.87959  9.878670  9.877812  9.877011
#> 13 26.899255 26.899255 27.019529 27.13876 27.249948 27.353646 27.450354
#> 14 23.844799 23.844799 23.985932 24.12584 24.256310 24.377991 24.491472
#> 15 19.141907 19.141907 19.094067 19.04664 19.002416 18.961170 18.922703
#> 16 17.178328 17.178328 16.993347 16.80998 16.638963 16.479477 16.330739
#> 17 17.154087 17.154087 17.095128 17.03668 16.982178 16.931346 16.883939
#> 18 25.420511 25.420511 25.301087 25.18270 25.072297 24.969333 24.873308
#> 19 28.460421 28.460421 28.344001 28.22859 28.120966 28.020591 27.926981
#> 20 20.426718 20.426718 20.158663 19.89294 19.645130 19.414019 19.198485
#> 21 22.317571 22.317571 22.356645 22.39538 22.431502 22.465191 22.496609
#>       0.9247    0.8624    0.8043    0.7501    0.6995    0.6524    0.6084
#> 1  22.408716 22.318925 22.235185 22.157089 22.084256 22.016332 21.952986
#> 2  21.428441 21.372234 21.319815 21.270928 21.225337 21.182818 21.143165
#> 3  24.711919 24.733444 24.753518 24.772239 24.789699 24.805982 24.821167
#> 4  20.806474 20.884899 20.958039 21.026250 21.089863 21.149189 21.204516
#> 5  18.748480 18.700057 18.654898 18.612782 18.573505 18.536875 18.502714
#> 6  21.154321 21.262473 21.363336 21.457400 21.545126 21.626938 21.703237
#> 7  22.531319 22.794643 23.040219 23.269245 23.482835 23.682029 23.867799
#> 8  19.518156 19.570684 19.619673 19.665360 19.707968 19.747704 19.784762
#> 9  17.610378 17.597095 17.584708 17.573156 17.562382 17.552334 17.542964
#> 10 17.536974 17.545859 17.554146 17.561874 17.569081 17.575803 17.582072
#> 11  9.665305  9.687176  9.707574  9.726596  9.744337  9.760882  9.776312
#> 12  9.876265  9.875568  9.874919  9.874314  9.873749  9.873222  9.872731
#> 13 27.540545 27.624658 27.703101 27.776258 27.844484 27.908112 27.967452
#> 14 24.597304 24.696003 24.788051 24.873895 24.953953 25.028615 25.098246
#> 15 18.886829 18.853373 18.822171 18.793073 18.765935 18.740627 18.717024
#> 16 16.192025 16.062660 15.942014 15.829500 15.724568 15.626708 15.535444
#> 17 16.839728 16.798496 16.760043 16.724182 16.690738 16.659548 16.630459
#> 18 24.783754 24.700237 24.622348 24.549708 24.481964 24.418786 24.359866
#> 19 27.839680 27.758263 27.682333 27.611521 27.545481 27.483892 27.426454
#> 20 18.997477 18.810016 18.635189 18.472145 18.320090 18.178282 18.046032
#> 21 22.525909 22.553235 22.578720 22.602486 22.624651 22.645322 22.664600
#>       0.5674    0.5292    0.4935    0.4602   0.4292   0.4003   0.3733   0.3482
#> 1  21.893909 21.838814 21.787432 21.761773 21.96340 22.15096 22.31480 22.04628
#> 2  21.106184 21.071695 21.039531 21.023788 21.35159 21.65651 21.92787 21.68454
#> 3  24.835329 24.848536 24.860854 24.867545 25.16125 25.43444 25.69036 25.96530
#> 4  21.256115 21.304236 21.349114 21.371499 21.17928 21.00048 20.84350 21.06900
#> 5  18.470855 18.441143 18.413434 18.398903 18.06849 17.76115 17.48540 17.64421
#> 6  21.774394 21.840755 21.902643 21.933692 21.78176 21.64042 21.51045 21.46237
#> 7  24.041048 24.202621 24.353305 24.429618 24.51374 24.59194 24.65777 24.45150
#> 8  19.819322 19.851553 19.881612 19.896541 19.72710 19.56949 19.41512 18.99814
#> 9  17.534225 17.526075 17.518474 17.514328 17.32253 17.14413 16.98040 16.93143
#> 10 17.587918 17.593370 17.598454 17.600800 17.45859 17.32632 17.20432 17.14531
#> 11  9.790701  9.804121  9.816637  9.823512 10.17019 10.49266 10.78787 10.84760
#> 12  9.872273  9.871846  9.871447  9.871701 10.15990 10.42797 10.67374 10.73736
#> 13 28.022792 28.074403 28.122535 28.147272 28.40266 28.64020 28.85925 28.96469
#> 14 25.163183 25.223744 25.280224 25.308188 24.93472 24.58731 24.27189 24.30884
#> 15 18.695012 18.674483 18.655338 18.645173 18.33720 18.05074 17.79359 17.93688
#> 16 15.450331 15.370954 15.296926 15.258955 14.91338 14.59195 14.29530 14.14337
#> 17 16.603332 16.578032 16.554438 16.542188 16.33857 16.14919 15.98304 16.22623
#> 18 24.304917 24.253672 24.205880 24.181836 24.25677 24.32649 24.39248 24.48846
#> 19 27.372887 27.322930 27.276340 27.252628 27.15236 27.05910 26.97295 26.92557
#> 20 17.922695 17.807671 17.700399 17.646486 17.84900 18.03739 18.20766 18.15702
#> 21 22.682579 22.699345 22.714982 22.723573 23.15753 23.56117 23.93223 24.06591
#>      0.3247   0.3028   0.2824   0.2634   0.2456   0.2291   0.2136   0.1992
#> 1  21.79628 21.56271 21.45504 20.66152 21.21342 21.21344 21.21344 21.21344
#> 2  21.45797 21.24632 21.14864 20.42799 20.76716 20.76711 20.76707 20.76704
#> 3  26.22118 26.46033 26.57018 26.40197 25.85317 25.85313 25.85310 25.85309
#> 4  21.27896 21.47511 21.56554 21.44143 21.73374 21.73373 21.73372 21.73372
#> 5  17.79211 17.93024 17.99410 17.87686 18.13694 18.13711 18.13724 18.13732
#> 6  21.41764 21.37582 21.35667 22.16611 24.09087 24.09052 24.09028 24.09012
#> 7  24.25945 24.08003 23.99733 24.52404 23.14225 23.14202 23.14186 23.14176
#> 8  18.61000 18.24730 18.08046 18.46701 19.37712 19.37696 19.37685 19.37677
#> 9  16.88588 16.84328 16.82383 16.47261 16.27842 16.27857 16.27867 16.27873
#> 10 17.09040 17.03907 17.01555 16.65228 16.19092 16.19105 16.19114 16.19120
#> 11 10.90314 10.95510 10.97874 11.10745 11.03685 11.03669 11.03658 11.03651
#> 12 10.79653 10.85187 10.87712 11.00382 10.71210 10.71200 10.71194 10.71191
#> 13 29.06279 29.15450 29.19650 28.67583 27.56622 27.56621 27.56621 27.56621
#> 14 24.34329 24.37543 24.39050 23.85073 23.44753 23.44768 23.44778 23.44784
#> 15 18.07033 18.19497 18.25259 18.10751 18.91737 18.91746 18.91753 18.91757
#> 16 14.00199 13.86982 13.80928 13.38825 12.88287 12.88324 12.88348 12.88365
#> 17 16.45265 16.66419 16.76173 16.83828 17.42818 17.42824 17.42829 17.42833
#> 18 24.57780 24.66128 24.69967 25.61406 27.06166 27.06147 27.06135 27.06127
#> 19 26.88149 26.84027 26.82139 27.41341 27.47984 27.47990 27.47994 27.47997
#> 20 18.10985 18.06580 18.04539 19.01525 17.52574 17.52592 17.52605 17.52614
#> 21 24.19028 24.30657 24.35975 24.09360 23.35762 23.35753 23.35747 23.35744
#>      0.1858   0.1733   0.1616   0.1507   0.1406   0.1311   0.1222   0.1140
#> 1  21.21343 21.21342 21.21341 21.21341 21.34643 22.42909 22.42905 22.11899
#> 2  20.76701 20.76699 20.76698 20.76696 21.06077 21.81778 21.81780 21.39119
#> 3  25.85308 25.85308 25.85308 25.85308 26.31676 25.22564 25.22576 25.47425
#> 4  21.73372 21.73372 21.73372 21.73372 21.62663 22.20006 22.19980 22.23706
#> 5  18.13737 18.13741 18.13743 18.13745 17.82682 18.87365 18.87339 18.77751
#> 6  24.09001 24.08994 24.08990 24.08987 23.39022 22.93903 22.93871 23.21354
#> 7  23.14169 23.14165 23.14162 23.14161 23.54583 24.49225 24.49237 24.24717
#> 8  19.37672 19.37668 19.37665 19.37664 18.90702 18.63134 18.63140 18.20639
#> 9  16.27877 16.27880 16.27881 16.27882 16.23654 15.42403 15.42424 14.96356
#> 10 16.19124 16.19126 16.19127 16.19128 16.26748 15.60112 15.60135 15.05338
#> 11 11.03646 11.03643 11.03641 11.03640 11.32862 11.69417 11.69409 11.97263
#> 12 10.71188 10.71187 10.71186 10.71185 10.99269 10.41434 10.41444 10.78147
#> 13 27.56620 27.56620 27.56620 27.56620 28.23131 28.92161 28.92167 28.62976
#> 14 23.44789 23.44791 23.44793 23.44794 23.44012 22.76678 22.76694 23.44523
#> 15 18.91760 18.91762 18.91763 18.91764 18.51387 18.61447 18.61425 18.16961
#> 16 12.88375 12.88382 12.88386 12.88389 12.71015 12.04825 12.04857 12.59621
#> 17 17.42835 17.42837 17.42838 17.42839 17.11562 18.20152 18.20113 18.23771
#> 18 27.06121 27.06118 27.06116 27.06115 26.52245 25.80940 25.80916 26.14705
#> 19 27.47999 27.48000 27.48001 27.48002 27.17439 27.23302 27.23300 27.52339
#> 20 17.52620 17.52625 17.52627 17.52629 17.66033 18.09215 18.09234 17.86757
#> 21 23.35742 23.35740 23.35739 23.35739 23.98592 22.77029 22.77054 23.14633
#>      0.1063   0.0992   0.0925   0.0862   0.0804   0.0750   0.0700   0.0652
#> 1  22.11904 22.11906 22.11907 22.11907 22.04644 22.04654 22.04662 22.04669
#> 2  21.39126 21.39130 21.39131 21.39131 21.33317 21.33328 21.33337 21.33344
#> 3  25.47415 25.47406 25.47400 25.47395 25.64275 25.64258 25.64242 25.64228
#> 4  22.23727 22.23748 22.23768 22.23788 22.35986 22.35996 22.36006 22.36017
#> 5  18.77780 18.77809 18.77836 18.77863 18.70697 18.70720 18.70743 18.70765
#> 6  23.21338 23.21322 23.21305 23.21289 23.23228 23.23222 23.23214 23.23206
#> 7  24.24733 24.24749 24.24764 24.24777 24.15321 24.15336 24.15350 24.15364
#> 8  18.20626 18.20611 18.20593 18.20573 18.04097 18.04104 18.04107 18.04106
#> 9  14.96356 14.96353 14.96349 14.96343 15.20515 15.20509 15.20503 15.20495
#> 10 15.05342 15.05344 15.05345 15.05343 15.29470 15.29469 15.29465 15.29461
#> 11 11.97262 11.97263 11.97265 11.97268 12.11865 12.11857 12.11851 12.11845
#> 12 10.78129 10.78113 10.78100 10.78088 10.80975 10.80957 10.80941 10.80927
#> 13 28.63001 28.63026 28.63050 28.63073 28.36638 28.36663 28.36688 28.36711
#> 14 23.44485 23.44451 23.44422 23.44396 23.36251 23.36220 23.36193 23.36167
#> 15 18.16988 18.17011 18.17031 18.17049 18.12973 18.12998 18.13021 18.13042
#> 16 12.59577 12.59537 12.59501 12.59469 12.31408 12.31381 12.31357 12.31333
#> 17 18.23803 18.23833 18.23864 18.23893 18.17263 18.17286 18.17310 18.17334
#> 18 26.14693 26.14682 26.14670 26.14660 25.95916 25.95917 25.95916 25.95916
#> 19 27.52331 27.52324 27.52318 27.52313 28.06408 28.06381 28.06356 28.06334
#> 20 17.86778 17.86799 17.86819 17.86839 17.75396 17.75412 17.75428 17.75445
#> 21 23.14606 23.14582 23.14561 23.14543 23.13356 23.13332 23.13310 23.13290
#>      0.0608   0.0567   0.0529   0.0494   0.0460   0.0429   0.0400   0.0373
#> 1  22.04674 22.04677 22.38105 22.38118 22.38130 22.38141 22.38152 22.38163
#> 2  21.33349 21.33353 21.55750 21.55753 21.55756 21.55758 21.55761 21.55764
#> 3  25.64215 25.64205 25.40432 25.40426 25.40420 25.40415 25.40409 25.40404
#> 4  22.36028 22.36039 22.36978 22.36981 22.36984 22.36986 22.36989 22.36992
#> 5  18.70788 18.70810 18.75759 18.75761 18.75763 18.75764 18.75766 18.75767
#> 6  23.23197 23.23188 23.17241 23.17231 23.17222 23.17213 23.17205 23.17197
#> 7  24.15377 24.15389 24.31374 24.31381 24.31388 24.31395 24.31401 24.31407
#> 8  18.04103 18.04098 18.05381 18.05364 18.05348 18.05333 18.05319 18.05305
#> 9  15.20487 15.20478 15.06172 15.06159 15.06146 15.06134 15.06123 15.06112
#> 10 15.29456 15.29450 15.16130 15.16117 15.16104 15.16092 15.16081 15.16070
#> 11 12.11841 12.11838 12.24649 12.24664 12.24678 12.24692 12.24705 12.24717
#> 12 10.80914 10.80902 10.76477 10.76482 10.76487 10.76492 10.76496 10.76501
#> 13 28.36735 28.36757 28.32850 28.32843 28.32836 28.32829 28.32822 28.32816
#> 14 23.36145 23.36124 23.44946 23.44969 23.44991 23.45012 23.45033 23.45052
#> 15 18.13061 18.13078 17.93072 17.93048 17.93024 17.93001 17.92979 17.92958
#> 16 12.31312 12.31292 12.42627 12.42648 12.42667 12.42686 12.42704 12.42721
#> 17 18.17359 18.17383 18.19231 18.19231 18.19231 18.19231 18.19231 18.19231
#> 18 25.95915 25.95913 25.71710 25.71689 25.71670 25.71651 25.71633 25.71615
#> 19 28.06313 28.06295 28.26607 28.26634 28.26659 28.26684 28.26708 28.26730
#> 20 17.75461 17.75477 17.68108 17.68103 17.68098 17.68093 17.68088 17.68083
#> 21 23.13271 23.13254 22.96400 22.96399 22.96397 22.96396 22.96395 22.96394
#>      0.0348   0.0325   0.0303   0.0282   0.0263   0.0246   0.0229   0.0214
#> 1  22.38173 22.38183 22.38192 22.38201 22.38209 22.38217 22.38225 22.38233
#> 2  21.55766 21.55769 21.55771 21.55773 21.55775 21.55777 21.55779 21.55781
#> 3  25.40400 25.40395 25.40391 25.40386 25.40382 25.40379 25.40375 25.40372
#> 4  22.36994 22.36996 22.36998 22.37000 22.37002 22.37004 22.37006 22.37007
#> 5  18.75768 18.75769 18.75771 18.75772 18.75772 18.75773 18.75774 18.75775
#> 6  23.17189 23.17182 23.17175 23.17168 23.17162 23.17156 23.17150 23.17144
#> 7  24.31413 24.31418 24.31424 24.31429 24.31433 24.31438 24.31442 24.31446
#> 8  18.05292 18.05279 18.05267 18.05255 18.05244 18.05233 18.05223 18.05213
#> 9  15.06102 15.06092 15.06082 15.06073 15.06064 15.06056 15.06048 15.06041
#> 10 15.16059 15.16049 15.16039 15.16030 15.16021 15.16013 15.16005 15.15997
#> 11 12.24729 12.24741 12.24751 12.24762 12.24772 12.24782 12.24791 12.24800
#> 12 10.76505 10.76509 10.76512 10.76516 10.76520 10.76523 10.76526 10.76529
#> 13 28.32810 28.32804 28.32798 28.32793 28.32787 28.32782 28.32778 28.32773
#> 14 23.45071 23.45089 23.45107 23.45123 23.45139 23.45155 23.45169 23.45183
#> 15 17.92938 17.92919 17.92901 17.92883 17.92866 17.92850 17.92834 17.92819
#> 16 12.42738 12.42754 12.42770 12.42784 12.42799 12.42812 12.42825 12.42838
#> 17 18.19231 18.19230 18.19230 18.19230 18.19229 18.19229 18.19229 18.19229
#> 18 25.71599 25.71583 25.71568 25.71553 25.71539 25.71526 25.71513 25.71500
#> 19 28.26752 28.26773 28.26793 28.26812 28.26831 28.26849 28.26866 28.26882
#> 20 17.68079 17.68074 17.68070 17.68066 17.68062 17.68059 17.68055 17.68052
#> 21 22.96393 22.96392 22.96391 22.96390 22.96389 22.96389 22.96388 22.96387
#>      0.0199   0.0186   0.0173   0.0162   0.0151   0.0141   0.0131   0.0122
#> 1  22.38240 22.38247 22.38253 22.38259 22.38265 22.38271 22.38276 22.38282
#> 2  21.55782 21.55784 21.55785 21.55787 21.55788 21.55790 21.55791 21.55792
#> 3  25.40368 25.40365 25.40362 25.40359 25.40357 25.40354 25.40351 25.40349
#> 4  22.37009 22.37010 22.37012 22.37013 22.37014 22.37016 22.37017 22.37018
#> 5  18.75776 18.75776 18.75777 18.75778 18.75778 18.75779 18.75779 18.75779
#> 6  23.17138 23.17133 23.17128 23.17124 23.17119 23.17115 23.17110 23.17106
#> 7  24.31450 24.31454 24.31458 24.31461 24.31465 24.31468 24.31471 24.31474
#> 8  18.05204 18.05195 18.05186 18.05178 18.05170 18.05162 18.05155 18.05148
#> 9  15.06033 15.06026 15.06020 15.06013 15.06007 15.06001 15.05996 15.05990
#> 10 15.15989 15.15982 15.15975 15.15969 15.15963 15.15957 15.15951 15.15945
#> 11 12.24808 12.24817 12.24824 12.24832 12.24839 12.24846 12.24852 12.24859
#> 12 10.76532 10.76535 10.76538 10.76541 10.76543 10.76546 10.76548 10.76551
#> 13 28.32768 28.32764 28.32760 28.32756 28.32752 28.32748 28.32745 28.32741
#> 14 23.45197 23.45210 23.45223 23.45235 23.45246 23.45257 23.45268 23.45278
#> 15 17.92804 17.92791 17.92777 17.92765 17.92753 17.92741 17.92730 17.92719
#> 16 12.42850 12.42861 12.42873 12.42883 12.42893 12.42903 12.42913 12.42922
#> 17 18.19228 18.19228 18.19228 18.19227 18.19227 18.19226 18.19226 18.19226
#> 18 25.71489 25.71477 25.71466 25.71456 25.71446 25.71437 25.71427 25.71419
#> 19 28.26898 28.26913 28.26927 28.26941 28.26954 28.26967 28.26979 28.26990
#> 20 17.68048 17.68045 17.68042 17.68039 17.68036 17.68034 17.68031 17.68029
#> 21 22.96387 22.96386 22.96386 22.96385 22.96385 22.96384 22.96384 22.96384
#>      0.0114   0.0106   0.0099   0.0092   0.0086   0.0080   0.0075   0.0070
#> 1  22.38287 22.38291 22.38296 22.38300 22.38304 22.38308 22.38312 22.38272
#> 2  21.55793 21.55794 21.55795 21.55796 21.55797 21.55798 21.55799 21.55756
#> 3  25.40347 25.40345 25.40343 25.40341 25.40339 25.40337 25.40335 25.40356
#> 4  22.37019 22.37020 22.37021 22.37022 22.37023 22.37024 22.37024 22.37079
#> 5  18.75780 18.75780 18.75781 18.75781 18.75781 18.75781 18.75782 18.75734
#> 6  23.17103 23.17099 23.17095 23.17092 23.17089 23.17086 23.17083 23.17133
#> 7  24.31476 24.31479 24.31481 24.31484 24.31486 24.31488 24.31490 24.31538
#> 8  18.05142 18.05135 18.05129 18.05123 18.05118 18.05112 18.05107 18.05200
#> 9  15.05985 15.05980 15.05976 15.05971 15.05967 15.05963 15.05959 15.05921
#> 10 15.15940 15.15935 15.15930 15.15926 15.15921 15.15917 15.15913 15.15875
#> 11 12.24865 12.24870 12.24876 12.24881 12.24886 12.24891 12.24896 12.24916
#> 12 10.76553 10.76555 10.76557 10.76559 10.76561 10.76562 10.76564 10.76579
#> 13 28.32738 28.32735 28.32732 28.32729 28.32726 28.32723 28.32721 28.32728
#> 14 23.45287 23.45297 23.45306 23.45314 23.45322 23.45330 23.45338 23.45379
#> 15 17.92709 17.92699 17.92690 17.92681 17.92672 17.92664 17.92656 17.92592
#> 16 12.42930 12.42939 12.42946 12.42954 12.42961 12.42968 12.42975 12.42980
#> 17 18.19225 18.19225 18.19225 18.19224 18.19224 18.19224 18.19223 18.19182
#> 18 25.71410 25.71402 25.71395 25.71387 25.71380 25.71374 25.71367 25.71260
#> 19 28.27001 28.27012 28.27022 28.27032 28.27041 28.27050 28.27059 28.27105
#> 20 17.68026 17.68024 17.68022 17.68020 17.68018 17.68016 17.68014 17.67989
#> 21 22.96383 22.96383 22.96383 22.96383 22.96382 22.96382 22.96382 22.96428
#>      0.0065   0.0061   0.0057   0.0053   0.0049   0.0046
#> 1  22.37589 22.37588 22.37587 22.37586 22.37586 22.37586
#> 2  21.54874 21.54873 21.54872 21.54871 21.54870 21.54870
#> 3  25.41436 25.41435 25.41433 25.41431 25.41429 25.41426
#> 4  22.39104 22.39123 22.39142 22.39160 22.39178 22.39195
#> 5  18.75117 18.75127 18.75137 18.75147 18.75157 18.75167
#> 6  23.17675 23.17688 23.17699 23.17709 23.17718 23.17726
#> 7  24.29198 24.29180 24.29163 24.29148 24.29133 24.29121
#> 8  18.08500 18.08531 18.08561 18.08589 18.08615 18.08640
#> 9  15.04986 15.04971 15.04956 15.04943 15.04929 15.04917
#> 10 15.14654 15.14637 15.14621 15.14606 15.14592 15.14578
#> 11 12.24702 12.24692 12.24684 12.24676 12.24668 12.24662
#> 12 10.76623 10.76615 10.76606 10.76597 10.76589 10.76581
#> 13 28.33277 28.33297 28.33315 28.33332 28.33349 28.33365
#> 14 23.44784 23.44759 23.44735 23.44712 23.44690 23.44668
#> 15 17.92287 17.92307 17.92326 17.92344 17.92361 17.92377
#> 16 12.43509 12.43496 12.43483 12.43470 12.43458 12.43445
#> 17 18.18859 18.18875 18.18891 18.18907 18.18922 18.18937
#> 18 25.68080 25.68077 25.68073 25.68069 25.68063 25.68057
#> 19 28.27907 28.27882 28.27859 28.27839 28.27820 28.27802
#> 20 17.68809 17.68819 17.68829 17.68839 17.68848 17.68858
#> 21 22.98029 22.98029 22.98029 22.98027 22.98025 22.98023
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 22.8 21.4 18.7 24.4 22.8 17.8 17.3 15.2 10.4 14.7 33.9 21.5 15.2
#> [16] 13.3 19.2 26.0 30.4 15.8 21.4
#> 
#> $X
#>               am       carb         cyl       disp       drat       gear
#>  [1,]  1.2747549  1.2565617 -0.05299989 -0.6026289  0.6067763  0.4677072
#>  [2,]  1.2747549  1.2565617 -0.05299989 -0.6026289  0.6067763  0.4677072
#>  [3,]  1.2747549 -1.3822179 -1.16599767 -1.0340658  0.4951974  0.4677072
#>  [4,] -0.7844645 -1.3822179 -0.05299989  0.2104638 -1.2231164 -0.9354143
#>  [5,] -0.7844645 -0.5026247  1.05999788  1.0567439 -1.0669061 -0.9354143
#>  [6,] -0.7844645 -0.5026247 -1.16599767 -0.7129772  0.1381452  0.4677072
#>  [7,] -0.7844645 -0.5026247 -1.16599767 -0.7619287  0.6514078  0.4677072
#>  [8,] -0.7844645  1.2565617 -0.05299989 -0.5395727  0.6514078  0.4677072
#>  [9,] -0.7844645  0.3769685  1.05999788  0.3581480 -1.2454322 -0.9354143
#> [10,] -0.7844645  0.3769685  1.05999788  0.3581480 -1.2454322 -0.9354143
#> [11,] -0.7844645  1.2565617  1.05999788  1.8864303 -1.4016425 -0.9354143
#> [12,] -0.7844645  1.2565617  1.05999788  1.7204931 -0.8883800 -0.9354143
#> [13,]  1.2747549 -1.3822179 -1.16599767 -1.3402201  1.3208807  0.4677072
#> [14,] -0.7844645 -1.3822179 -1.16599767 -0.9336738  0.1604610 -0.9354143
#> [15,] -0.7844645 -0.5026247  1.05999788  0.5921195 -1.0669061 -0.9354143
#> [16,] -0.7844645  1.2565617  1.05999788  0.9737753  0.2274083 -0.9354143
#> [17,] -0.7844645 -0.5026247  1.05999788  1.3886185 -1.2231164 -0.9354143
#> [18,]  1.2747549 -0.5026247 -1.16599767 -0.9320144  1.7895118  1.8708287
#> [19,]  1.2747549 -0.5026247 -1.16599767 -1.1410954  0.3166713  1.8708287
#> [20,]  1.2747549  1.2565617  1.05999788  0.9820721  1.3208807  1.8708287
#> [21,]  1.2747549 -0.5026247 -1.16599767 -0.9262066  1.0754073  0.4677072
#>               hp        qsec         vs           wt
#>  [1,] -0.5600325 -0.84830046 -0.8660254 -0.628639617
#>  [2,] -0.5600325 -0.53557276 -0.8660254 -0.360018076
#>  [3,] -0.8531873  0.35235050  1.1547005 -0.944664959
#>  [4,] -0.5600325  0.81585762  1.1547005 -0.001856022
#>  [5,]  0.5608536 -0.53557276 -0.8660254  0.235162985
#>  [6,] -1.3877638  1.12858531  1.1547005 -0.028191467
#>  [7,] -0.8186985  2.74806800  1.1547005 -0.070328179
#>  [8,] -0.3358553  0.51429877  1.1547005  0.235162985
#>  [9,]  0.6470756 -0.21167623 -0.8660254  0.540654149
#> [10,]  0.6470756  0.01170070 -0.8660254  0.593325039
#> [11,]  1.2506297 -0.08881892 -0.8660254  2.325143914
#> [12,]  1.5092957 -0.31219584 -0.8660254  2.241923907
#> [13,] -1.3360306  1.07274108  1.1547005 -1.455572596
#> [14,] -0.7842097  1.13416973  1.1547005 -0.791919377
#> [15,]  0.1297436 -0.37920892 -0.8660254  0.229895896
#> [16,]  1.7679618 -1.43466488 -0.8660254  0.656530108
#> [17,]  0.5608536 -0.51881950 -0.8660254  0.661797197
#> [18,] -0.8876761 -0.71427430 -0.8660254 -1.134280164
#> [19,] -0.5082993 -0.60258584  1.1547005 -1.794773130
#> [20,]  2.0956054 -1.94284738 -0.8660254 -0.049259823
#> [21,] -0.5772769  0.34676608  1.1547005 -0.460092768
#> attr(,"center")
#>  [1]   0.3809524   2.5714286   6.0952381 232.6333333   3.6280952   3.6666667
#>  [7] 142.4761905  17.9790476   0.4285714   3.2167619
#> attr(,"scale")
#>  [1]   0.4856209   1.1368892   1.7969488 120.5274651   0.4481137   0.7126966
#>  [7]  57.9898341   1.7906953   0.4948717   0.9492910
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 28
#> 
#> $lambda.min
#> [1] 0.6995372
#> 
#> $null.dev
#> [1] 29.66952
#> 
#> $Bias
#> [1] 5.156805
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 6.650804