Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 37.98096 36.68389 33.60199 30.30329 27.65649 25.56171 23.93308 22.69685
#>   [9] 21.78976 21.15761 20.75402 20.53937 20.47986 20.54670 20.71546 20.88003
#>  [17] 20.66786 20.45391 20.26950 19.91709 19.17224 17.86583 16.01760 14.65424
#>  [25] 13.84214 13.36026 13.10824 13.26814 15.28820 14.21375 13.94979 13.71947
#>  [33] 13.54894 14.90438 14.87769 17.31059 19.47443 19.18541 19.67398 21.33651
#>  [41] 21.23159 21.48672 37.66890 37.25139 38.12744 38.15336 41.46907 41.39893
#>  [49] 41.34259 41.74820 42.76860 42.94200 40.41518 41.22905 42.07319 44.59040
#>  [57] 44.59394 44.59757 44.98035 44.98367 46.16494 46.11304 46.87672 46.88004
#>  [65] 46.46778 46.48836 46.48918 46.59574 47.78595 47.78699 47.78808 47.79066
#>  [73] 47.56955 47.56832 47.56715 47.56602 47.56493 47.65724 47.65630 47.65539
#>  [81] 47.65453 47.65370 47.65290 47.69661 47.71273 50.93252 50.88200 50.88302
#>  [89] 50.88402 50.83084 50.83181 50.83276 50.83370 50.83463 50.83555 50.83647
#>  [97] 50.83738 50.84661 50.84749 50.84837
#> 
#> $cvse
#>   [1] 11.660404 11.596297 10.772894  9.706294  8.754150  7.896795  7.119648
#>   [8]  6.412384  5.768243  5.183503  4.657100  4.190340  3.786608  3.450853
#>  [15]  3.188612  3.011860  2.889310  2.847076  2.855634  2.829341  2.750951
#>  [22]  2.778156  2.804131  2.872798  2.907492  2.889778  2.835887  2.749602
#>  [29]  3.032518  3.002356  3.005718  3.010256  2.999815  3.239542  3.275701
#>  [36]  4.983283  5.250265  5.215954  5.170250  5.791061  5.699347  5.632343
#>  [43] 20.850106 20.856509 20.964724 20.948487 21.703660 21.708535 21.712921
#>  [50] 21.686636 21.670110 21.656583 19.391206 19.355827 19.313125 19.508951
#>  [57] 19.511383 19.513665 19.498566 19.500616 20.544246 20.555643 20.505546
#>  [64] 20.507586 20.394507 20.390784 20.392174 20.589846 20.822731 20.824265
#>  [71] 20.825763 20.827056 20.745546 20.744954 20.744365 20.743779 20.743198
#>  [78] 20.733561 20.732984 20.732412 20.731848 20.731290 20.730740 20.727983
#>  [85] 20.725859 21.921567 21.925025 21.925163 21.925297 21.927665 21.927796
#>  [92] 21.927924 21.928052 21.928180 21.928308 21.928436 21.928566 21.927803
#>  [99] 21.927932 21.928063
#> 
#> $fold
#>  [1]  6  1  2  1  4  1  4 10 10  9  5  7  2  5  6  3  9  8  3  7  8
#> 
#> $lambda
#>   [1] 5.028863324 4.689934767 4.373848861 4.079066086 3.804150684 3.547763660
#>   [7] 3.308656263 3.085663905 2.877700485 2.683753104 2.502877127 2.334191585
#>  [13] 2.176874885 2.030160803 1.893334760 1.765730334 1.646726019 1.535742197
#>  [19] 1.432238313 1.335710244 1.245687844 1.161732652 1.083435760 1.010415816
#>  [25] 0.942317171 0.878808148 0.819579420 0.764342510 0.712828383 0.664786136
#>  [31] 0.619981775 0.578197079 0.539228531 0.502886332 0.468993477 0.437384886
#>  [37] 0.407906609 0.380415068 0.354776366 0.330865626 0.308566390 0.287770049
#>  [43] 0.268375310 0.250287713 0.233419159 0.217687488 0.203016080 0.189333476
#>  [49] 0.176573033 0.164672602 0.153574220 0.143223831 0.133571024 0.124568784
#>  [55] 0.116173265 0.108343576 0.101041582 0.094231717 0.087880815 0.081957942
#>  [61] 0.076434251 0.071282838 0.066478614 0.061998178 0.057819708 0.053922853
#>  [67] 0.050288633 0.046899348 0.043738489 0.040790661 0.038041507 0.035477637
#>  [73] 0.033086563 0.030856639 0.028777005 0.026837531 0.025028771 0.023341916
#>  [79] 0.021768749 0.020301608 0.018933348 0.017657303 0.016467260 0.015357422
#>  [85] 0.014322383 0.013357102 0.012456878 0.011617327 0.010834358 0.010104158
#>  [91] 0.009423172 0.008788081 0.008195794 0.007643425 0.007128284 0.006647861
#>  [97] 0.006199818 0.005781971 0.005392285 0.005028863
#> 
#> $fit
#> $beta
#>               5.0289     4.6899    4.3738    4.0791    3.8042    3.5478
#> (Intercept) 20.63333 22.3598459 23.969997 25.471630 26.872057 28.178101
#> am           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000 -0.5348867 -1.033724 -1.498941 -1.932804 -2.337426
#>                3.3087    3.0857   2.8777    2.6838    2.5029    2.3342
#> (Intercept) 29.396121 30.532051 31.59142 32.579397 33.500785 34.360074
#> am           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> wt          -2.714778 -3.066698 -3.39490 -3.700982 -3.986435 -4.252649
#>                2.1769    2.0302    1.8933      1.7657      1.6467     1.5357
#> (Intercept) 35.161450 35.908817 36.605813 36.98691538 35.95068438 34.1293436
#> am           0.000000  0.000000  0.000000  0.00000000  0.00000000  0.0000000
#> carb         0.000000  0.000000  0.000000  0.00000000  0.00000000  0.0000000
#> cyl          0.000000  0.000000  0.000000  0.00000000  0.00000000  0.0000000
#> disp         0.000000  0.000000  0.000000  0.00000000  0.00000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.00000000  0.00000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.00000000  0.00000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.00000000  0.00000000  0.0000000
#> qsec         0.000000  0.000000  0.000000  0.01343273  0.09088608  0.1848985
#> vs           0.000000  0.000000  0.000000  0.00000000  0.00000000  0.0000000
#> wt          -4.500921 -4.732461 -4.948396 -5.14170940 -5.25454116 -5.2168977
#>                 1.4322     1.3357     1.2457     1.1617     1.0834     1.0104
#> (Intercept) 32.4310224 30.8467989 29.3693467 27.9914698 26.7064572 25.5080501
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.2725611  0.3543342  0.4305961  0.5017182  0.5680469  0.6299053
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.1817967 -5.1490540 -5.1185180 -5.0900400 -5.0634813 -5.0387126
#>                 0.9423     0.8788     0.8196     0.7643     0.7128     0.6648
#> (Intercept) 24.3904117 23.3480984 22.3760335 21.4694825 20.6240300 19.8355582
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.6875946  0.7413959  0.7915711  0.8383647  0.8820046  0.9227033
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.0156133 -4.9940707 -4.9739801 -4.9552435 -4.9377696 -4.9214735
#>                0.6200     0.5782     0.5392     0.5029    0.4690    0.4374
#> (Intercept) 19.100227 18.9278906 17.3890998 14.6855885 12.159521  9.803661
#> am           0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> carb         0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> cyl          0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> disp         0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> drat         0.000000  0.0000000  0.2783649  0.7671955  1.223909  1.649849
#> gear         0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> hp           0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> qsec         0.960659  0.9695545  0.9760186  0.9874357  0.998112  1.008069
#> vs           0.000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> wt          -4.906276 -4.9027138 -4.7746685 -4.5497845 -4.339672 -4.143717
#>                 0.4079     0.3804     0.3548     0.3309     0.3086     0.2878
#> (Intercept)  7.7848036  8.1313198  8.4064611  8.4040908  8.4023513  8.4010752
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb        -0.2358975 -0.4753092 -0.6641456 -0.6643367 -0.6644768 -0.6645795
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         2.4161402  2.7606959  3.0321797  3.0328755  3.0333859  3.0337602
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.9142971  0.8090179  0.7259592  0.7259015  0.7258593  0.7258282
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -3.6652475 -3.3789165 -3.1532086 -3.1527799 -3.1524655 -3.1522349
#>                 0.2684     0.2503     0.2334    0.2177    0.2030    0.1893
#> (Intercept)  8.4001392  8.3994527  5.4321679  9.155166  9.154009  9.153559
#> am           0.0000000  0.0000000  2.1979482  2.636603  2.635880  2.635238
#> carb        -0.6646549 -0.6647101 -0.7538138  0.000000  0.000000  0.000000
#> cyl          0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> disp         0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> drat         3.0340348  3.0342362  2.2228560  3.028948  3.029613  3.030063
#> gear         0.0000000  0.0000000  0.0000000 -2.258539 -2.258535 -2.258498
#> hp           0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> qsec         0.7258055  0.7257888  0.9128264  1.046183  1.046127  1.046075
#> vs           0.0000000  0.0000000  0.0000000  0.000000  0.000000  0.000000
#> wt          -3.1520657 -3.1519416 -2.5904111 -3.455539 -3.455520 -3.455553
#>                0.1766    0.1647    0.1536     0.1432     0.1336     0.1246
#> (Intercept)  3.361546  3.357511  3.354276  2.9967573  3.0000543  3.0028127
#> am           3.006841  3.007733  3.008445  3.0113545  3.0109097  3.0105281
#> carb         0.000000  0.000000  0.000000 -0.3446616 -0.3448223 -0.3449523
#> cyl          0.000000  0.000000  0.000000  0.0000000  0.0000000  0.0000000
#> disp         0.000000  0.000000  0.000000  0.0000000  0.0000000  0.0000000
#> drat         2.974165  2.973877  2.973652  2.9788505  2.9788332  2.9788301
#> gear        -2.169691 -2.169599 -2.169527 -1.6520695 -1.6518056 -1.6515957
#> hp           0.000000  0.000000  0.000000  0.0000000  0.0000000  0.0000000
#> qsec         1.424115  1.424331  1.424503  1.3028127  1.3025736  1.3023745
#> vs          -1.702739 -1.703161 -1.703494 -1.5671572 -1.5665237 -1.5659993
#> wt          -3.616361 -3.616162 -3.615999 -3.1707790 -3.1706519 -3.1705525
#>                 0.1162     0.1083     0.1010     0.0942    0.0879     0.0820
#> (Intercept)  3.0051187  3.0070454  3.0086543  3.0099971  3.011118  3.0120517
#> am           3.0102014  3.0099223  3.0096843  3.0094819  3.009310  3.0091645
#> carb        -0.3450576 -0.3451429 -0.3452119 -0.3452678 -0.345313 -0.3453496
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> drat         2.9788361  2.9788476  2.9788620  2.9788775  2.978893  2.9789080
#> gear        -1.6514286 -1.6512952 -1.6511887 -1.6511035 -1.651035 -1.6509807
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> qsec         1.3022089  1.3020712  1.3019567  1.3018617  1.301783  1.3017175
#> vs          -1.5655655 -1.5652069 -1.5649106 -1.5646659 -1.564464 -1.5642974
#> wt          -3.1704749 -3.1704146 -3.1703679 -3.1703319 -3.170304 -3.1702833
#>                  0.0764      0.0713      0.0665      0.0620      0.0578
#> (Intercept) 17.41741656 17.41809042 17.41833544 17.41821116 17.41777161
#> am           2.28637274  2.28595623  2.28559381  2.28528072  2.28501247
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl         -1.09965135 -1.09953931 -1.09941548 -1.09928260 -1.09914306
#> disp         0.01250768  0.01250316  0.01249897  0.01249509  0.01249152
#> drat         2.50515102  2.50548073  2.50579224  2.50608557  2.50636090
#> gear        -2.20366394 -2.20376577 -2.20385267 -2.20392606 -2.20398727
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         1.05629547  1.05619205  1.05610833  1.05604217  1.05599161
#> vs          -1.64789425 -1.64784203 -1.64778739 -1.64773110 -1.64767389
#> wt          -4.08188654 -4.08160724 -4.08135563 -4.08112981 -4.08092797
#>                  0.0539      0.0503      0.0469      0.0437     0.0408
#> (Intercept) 17.41706572 17.41613758 17.41502680 17.41376883 17.4123952
#> am           2.28478485  2.28459390  2.28443596  2.28430760  2.2842056
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> cyl         -1.09899897 -1.09885219 -1.09870432 -1.09855674 -1.0984106
#> disp         0.01248824  0.01248523  0.01248248  0.01247997  0.0124777
#> drat         2.50661853  2.50685886  2.50708238  2.50728965  2.5074813
#> gear        -2.20403754 -2.20407802 -2.20410978 -2.20413381 -2.2041510
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.0000000
#> qsec         1.05595483  1.05593014  1.05591603  1.05591108  1.0559140
#> vs          -1.64761634 -1.64755899 -1.64750228 -1.64744659 -1.6473922
#> wt          -4.08074835 -4.08058924 -4.08044902 -4.08032616 -4.0802192
#>                  0.0380      0.0355      0.0331     0.0309      0.0288
#> (Intercept) 19.37016856 19.37482902 19.37927794 19.3835247 19.38757845
#> am           2.19314829  2.19288584  2.19263511  2.1923956  2.19216680
#> carb         0.15735851  0.15751541  0.15766517  0.1578081  0.15794452
#> cyl         -1.25251377 -1.25283143 -1.25313462 -1.2534240 -1.25370015
#> disp         0.01451845  0.01452129  0.01452401  0.0145266  0.01452907
#> drat         2.44060498  2.44050015  2.44040024  2.4403050  2.44021422
#> gear        -2.43775567 -2.43804532 -2.43832182 -2.4385858 -2.43883770
#> hp           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> qsec         1.06723515  1.06716797  1.06710380  1.0670425  1.06698397
#> vs          -1.69475705 -1.69486497 -1.69496794 -1.6950662 -1.69515996
#> wt          -4.36856794 -4.36886373 -4.36914600 -4.3694154 -4.36967246
#>                  0.0268      0.0250      0.0233      0.0218      0.0203
#> (Intercept) 19.39144773 19.39514087 19.39866579 19.40203007 19.40524098
#> am           2.19194829  2.19173962  2.19154035  2.19135009  2.19116842
#> carb         0.15807472  0.15819898  0.15831758  0.15843076  0.15853877
#> cyl         -1.25396373 -1.25421527 -1.25445534 -1.25468444 -1.25490309
#> disp         0.01453142  0.01453367  0.01453582  0.01453787  0.01453982
#> drat         2.44012767  2.44004514  2.43996645  2.43989140  2.43981983
#> gear        -2.43907817 -2.43930770 -2.43952678 -2.43973587 -2.43993543
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         1.06692808  1.06687470  1.06682374  1.06677509  1.06672864
#> vs          -1.69524942 -1.69533479 -1.69541624 -1.69549397 -1.69556813
#> wt          -4.36991780 -4.37015192 -4.37037536 -4.37058858 -4.37079206
#>                  0.0189      0.0177      0.0165      0.0154      0.0143
#> (Intercept) 19.40830545 19.41123011 19.41402131 19.41668510 19.41922727
#> am           2.19099499  2.19082941  2.19067135  2.19052046  2.19037644
#> carb         0.15864185  0.15874023  0.15883411  0.15892370  0.15900920
#> cyl         -1.25511174 -1.25531087 -1.25550090 -1.25568224 -1.25585530
#> disp         0.01454169  0.01454346  0.01454516  0.01454678  0.01454833
#> drat         2.43975157  2.43968646  2.43962435  2.43956510  2.43950858
#> gear        -2.44012588 -2.44030765 -2.44048112 -2.44064668 -2.44080467
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         1.06668430  1.06664197  1.06660157  1.06656300  1.06652619
#> vs          -1.69563890 -1.69570642 -1.69577086 -1.69583234 -1.69589101
#> wt          -4.37098625 -4.37117156 -4.37134841 -4.37151717 -4.37167823
#>                  0.0134      0.0125      0.0116      0.0108      0.0101
#> (Intercept) 19.42165335 19.42396863 19.42617814 19.42828672 19.43029896
#> am           2.19023896  2.19010775  2.18998251  2.18986297  2.18974888
#> carb         0.15909079  0.15916866  0.15924297  0.15931388  0.15938155
#> cyl         -1.25602045 -1.25617805 -1.25632845 -1.25647198 -1.25660894
#> disp         0.01454981  0.01455121  0.01455256  0.01455384  0.01455506
#> drat         2.43945466  2.43940322  2.43935414  2.43930731  2.43926263
#> gear        -2.44095545 -2.44109934 -2.44123666 -2.44136770 -2.44149276
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         1.06649106  1.06645753  1.06642552  1.06639498  1.06636582
#> vs          -1.69594699 -1.69600040 -1.69605138 -1.69610001 -1.69614643
#> wt          -4.37183192 -4.37197859 -4.37211855 -4.37225212 -4.37237959
#>                  0.0094      0.0088      0.0082       0.0076       0.0071
#> (Intercept) 19.43221925 19.43405179 19.43580058 20.824966877 20.827873005
#> am           2.18964000  2.18953608  2.18943691  2.009271056  2.008730222
#> carb         0.15944612  0.15950775  0.15956656  0.132928183  0.132797073
#> cyl         -1.25673965 -1.25686438 -1.25698340 -1.353108084 -1.353224192
#> disp         0.01455623  0.01455734  0.01455841  0.013812995  0.013808246
#> drat         2.43922000  2.43917932  2.43914051  2.509657291  2.510001047
#> gear        -2.44161210 -2.44172599 -2.44183467 -2.594546461 -2.594848837
#> hp           0.00000000  0.00000000  0.00000000  0.003282895  0.003289882
#> qsec         1.06633800  1.06631145  1.06628611  1.041051770  1.040930913
#> vs          -1.69619071 -1.69623298 -1.69627330 -1.766875469 -1.767041602
#> wt          -4.37250123 -4.37261731 -4.37272809 -4.417524726 -4.417328510
#>                   0.0066       0.0062      0.0058       0.0054       0.0050
#> (Intercept) 20.830644066 20.833284477 20.83579854 20.838190471 20.840464349
#> am           2.008208337  2.007704899  2.00721942  2.006751401  2.006300382
#> carb         0.132668711  0.132543071  0.13242013  0.132299846  0.132182203
#> cyl         -1.353333063 -1.353434970 -1.35353018 -1.353618937 -1.353701503
#> disp         0.013803611  0.013799087  0.01379467  0.013790367  0.013786168
#> drat         2.510335621  2.510661185  2.51097791  2.511285966  2.511585520
#> gear        -2.595138383 -2.595415495 -2.59568056 -2.595933959 -2.596176058
#> hp           0.003296624  0.003303128  0.00330940  0.003315447  0.003321275
#> qsec         1.040814052  1.040701087  1.04059192  1.040486453  1.040384592
#> vs          -1.767201255 -1.767354616 -1.76750187 -1.767643196 -1.767778772
#> wt          -4.417134111 -4.416941608 -4.41675108 -4.416562579 -4.416376183
#> 
#> $iter
#>   [1]   0   9   5   4   3   3   3   2   2   2   2   2   2   2   2   5   4   4
#>  [19]   3   3   3   3   3   3   3   3   3   3   3   3   3   3  22  23  23  23
#>  [37]  72  50  27   1   1   1   1   1  33  68   1   1  38   1   1  56   1   1
#>  [55]   1   1   1   1   1   1 138   1   1   1   1   1   1   1   1   1  94   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1 102   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.028863324 4.689934767 4.373848861 4.079066086 3.804150684 3.547763660
#>   [7] 3.308656263 3.085663905 2.877700485 2.683753104 2.502877127 2.334191585
#>  [13] 2.176874885 2.030160803 1.893334760 1.765730334 1.646726019 1.535742197
#>  [19] 1.432238313 1.335710244 1.245687844 1.161732652 1.083435760 1.010415816
#>  [25] 0.942317171 0.878808148 0.819579420 0.764342510 0.712828383 0.664786136
#>  [31] 0.619981775 0.578197079 0.539228531 0.502886332 0.468993477 0.437384886
#>  [37] 0.407906609 0.380415068 0.354776366 0.330865626 0.308566390 0.287770049
#>  [43] 0.268375310 0.250287713 0.233419159 0.217687488 0.203016080 0.189333476
#>  [49] 0.176573033 0.164672602 0.153574220 0.143223831 0.133571024 0.124568784
#>  [55] 0.116173265 0.108343576 0.101041582 0.094231717 0.087880815 0.081957942
#>  [61] 0.076434251 0.071282838 0.066478614 0.061998178 0.057819708 0.053922853
#>  [67] 0.050288633 0.046899348 0.043738489 0.040790661 0.038041507 0.035477637
#>  [73] 0.033086563 0.030856639 0.028777005 0.026837531 0.025028771 0.023341916
#>  [79] 0.021768749 0.020301608 0.018933348 0.017657303 0.016467260 0.015357422
#>  [85] 0.014322383 0.013357102 0.012456878 0.011617327 0.010834358 0.010104158
#>  [91] 0.009423172 0.008788081 0.008195794 0.007643425 0.007128284 0.006647861
#>  [97] 0.006199818 0.005781971 0.005392285 0.005028863
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 32
#> 
#> $loss
#>   [1] 752.5867 650.6356 565.3384 494.2984 435.4465 386.9973 347.4115 315.3625
#>   [9] 289.7082 269.4660 253.7914 241.9594 233.3481 227.4248 223.7338 220.0471
#>  [17] 209.4729 198.2693 188.9635 181.2755 174.9692 169.8390 165.7077 162.4229
#>  [25] 159.8536 157.8873 156.4273 155.3907 154.7068 154.3149 154.1633 154.1576
#>  [33] 150.5773 145.5339 142.2542 140.4425 134.0825 130.7816 129.9807 129.9807
#>  [41] 129.9807 129.9807 129.9807 129.9807 124.4285 121.0427 121.0427 121.0427
#>  [49] 117.7432 117.7432 117.7432 116.0380 116.0380 116.0380 116.0380 116.0380
#>  [57] 116.0380 116.0380 116.0380 116.0380 111.2608 111.2608 111.2608 111.2608
#>  [65] 111.2608 111.2608 111.2607 111.2607 111.2607 111.2607 111.0491 111.0491
#>  [73] 111.0491 111.0491 111.0491 111.0491 111.0491 111.0491 111.0491 111.0491
#>  [81] 111.0490 111.0490 111.0490 111.0490 111.0490 111.0490 111.0490 111.0490
#>  [89] 111.0490 111.0490 111.0490 111.0490 111.0490 111.0078 111.0078 111.0078
#>  [97] 111.0078 111.0077 111.0077 111.0077
#> 
#> $linear.predictors
#>      5.0289   4.6899   4.3738   4.0791   3.8042   3.5478   3.3087   3.0857
#> 1  20.63333 20.95844 21.26164 21.54440 21.80811 22.05404 22.28340 22.49730
#> 2  20.63333 21.11891 21.57176 21.99409 22.38795 22.75527 23.09784 23.41731
#> 3  20.63333 20.64018 20.64657 20.65253 20.65809 20.66327 20.66811 20.67262
#> 4  20.63333 20.51984 20.41399 20.31527 20.22321 20.13735 20.05728 19.98261
#> 5  20.63333 20.50914 20.39331 20.28529 20.18455 20.09061 20.00299 19.92128
#> 6  20.63333 20.65356 20.67242 20.69001 20.70641 20.72171 20.73598 20.74928
#> 7  20.63333 20.67495 20.71377 20.74997 20.78372 20.81521 20.84457 20.87195
#> 8  20.63333 20.51984 20.41399 20.31527 20.22321 20.13735 20.05728 19.98261
#> 9  20.63333 20.18286 19.76274 19.37094 19.00554 18.66478 18.34697 18.05059
#> 10 20.63333 19.45862 18.36308 17.34137 16.38853 15.49990 14.67116 13.89828
#> 11 20.63333 19.50088 18.44474 17.45979 16.54122 15.68456 14.88563 14.14055
#> 12 20.63333 21.18310 21.69580 22.17396 22.61989 23.03576 23.42361 23.78532
#> 13 20.63333 21.49600 22.30053 23.05084 23.75058 24.40316 25.01175 25.57933
#> 14 20.63333 21.37833 22.07311 22.72107 23.32536 23.88892 24.41450 24.90466
#> 15 20.63333 20.47704 20.33129 20.19536 20.06859 19.95036 19.84010 19.73727
#> 16 20.63333 20.52251 20.41916 20.32277 20.23287 20.14904 20.07086 19.99794
#> 17 20.63333 20.30321 19.99533 19.70820 19.44042 19.19070 18.95780 18.74060
#> 18 20.63333 21.21519 21.75783 22.26390 22.73586 23.17601 23.58650 23.96932
#> 19 20.63333 20.66425 20.69309 20.71999 20.74507 20.76846 20.79027 20.81062
#> 20 20.63333 20.45030 20.27960 20.12041 19.97195 19.83349 19.70436 19.58394
#> 21 20.63333 20.87286 21.09624 21.30457 21.49886 21.68006 21.84904 22.00663
#>      2.8777   2.6838   2.5029   2.3342   2.1769   2.0302    1.8933    1.7657
#> 1  22.69679 22.88283 23.05633 23.21813 23.36904 23.50977 23.641015 23.736739
#> 2  23.71526 23.99312 24.25226 24.49393 24.71931 24.92951 25.125534 25.308133
#> 3  20.67682 20.68074 20.68440 20.68781 20.69099 20.69395 20.696720 20.717452
#> 4  19.91297 19.84802 19.78745 19.73096 19.67828 19.62915 19.583331 19.528060
#> 5  19.84507 19.77400 19.70772 19.64591 19.58826 19.53450 19.484363 19.468211
#> 6  20.76169 20.77327 20.78406 20.79412 20.80351 20.81227 20.820430 20.853517
#> 7  20.89749 20.92130 20.94352 20.96423 20.98355 21.00156 21.018366 21.098140
#> 8  19.91297 19.84802 19.78745 19.73096 19.67828 19.62915 19.583331 19.553314
#> 9  17.77418 17.51640 17.27600 17.05179 16.84270 16.64770 16.465841 16.293888
#> 10 13.17749 12.50527 11.87836 11.29371 10.74845 10.23995  9.765713  9.337655
#> 11 13.44568 12.79765 12.19329 11.62967 11.10403 10.61381 10.156637  9.738477
#> 12 24.12264 24.43724 24.73063 25.00425 25.25942 25.49740 25.719342 25.936690
#> 13 26.10866 26.60231 27.06269 27.49205 27.89246 28.26589 28.614153 28.931829
#> 14 25.36178 25.78810 26.18568 26.55646 26.90226 27.22475 27.525506 27.819190
#> 15 19.64138 19.55194 19.46854 19.39075 19.31821 19.25055 19.187459 19.114708
#> 16 19.92994 19.86653 19.80738 19.75222 19.70079 19.65281 19.608073 19.557530
#> 17 18.53803 18.34912 18.17294 18.00864 17.85541 17.71250 17.579231 17.446071
#> 18 24.32634 24.65930 24.96981 25.25941 25.52948 25.78135 26.016245 26.207984
#> 19 20.82959 20.84729 20.86379 20.87918 20.89353 20.90691 20.919398 20.882471
#> 20 19.47163 19.36689 19.26921 19.17812 19.09316 19.01393 18.940039 18.827131
#> 21 22.15360 22.29067 22.41850 22.53771 22.64889 22.75257 22.849272 22.942812
#>       1.6467    1.5357    1.4322    1.3357    1.2457    1.1617    1.0834
#> 1  23.679771 23.504502 23.341070 23.188619 23.046442 22.913847 22.790188
#> 2  25.451539 25.467103 25.481616 25.495153 25.507779 25.519553 25.530534
#> 3  20.824160 20.951445 21.070133 21.180847 21.284100 21.380393 21.470197
#> 4  19.421944 19.330189 19.244631 19.164821 19.090391 19.020976 18.956240
#> 5  19.607689 19.817526 20.013191 20.195710 20.365928 20.524674 20.672721
#> 6  21.006420 21.185411 21.352312 21.508001 21.653196 21.788606 21.914890
#> 7  21.480171 21.930293 22.350011 22.741532 23.106666 23.447191 23.764766
#> 8  19.592810 19.677798 19.757046 19.830970 19.899911 19.964206 20.024168
#> 9  16.146120 16.113805 16.083672 16.055564 16.029351 16.004904 15.982105
#> 10  9.069643  9.127783  9.181995  9.232566  9.279728  9.323711  9.364731
#> 11  9.448397  9.465958  9.482333  9.497607  9.511852  9.525137  9.537527
#> 12 26.160246 26.252143 26.337834 26.417767 26.492313 26.561835 26.626672
#> 13 29.147811 29.128375 29.110252 29.093346 29.077580 29.062876 29.049164
#> 14 28.117234 28.235817 28.346391 28.449536 28.545729 28.635439 28.719103
#> 15 18.987948 18.885102 18.789203 18.699747 18.616320 18.538515 18.465955
#> 16 19.473665 19.408045 19.346857 19.289780 19.236550 19.186907 19.140611
#> 17 17.296581 17.222892 17.154180 17.090085 17.030309 16.974562 16.922572
#> 18 26.223764 26.052988 25.893747 25.745205 25.606673 25.477478 25.356991
#> 19 20.611637 20.272807 19.956862 19.662144 19.387288 19.130957 18.891902
#> 20 18.518909 18.204538 17.911400 17.637956 17.382941 17.145113 16.923314
#> 21 23.033541 23.065481 23.095263 23.123045 23.148954 23.173117 23.195652
#>       1.0104    0.9423    0.8788    0.8196    0.7643    0.7128    0.6648
#> 1  22.674864 22.567313 22.467010 22.373467 22.286228 22.204869 22.128994
#> 2  25.540775 25.550325 25.559232 25.567539 25.575285 25.582510 25.589248
#> 3  21.553948 21.632055 21.704897 21.772831 21.836185 21.895270 21.950373
#> 4  18.895867 18.839563 18.787053 18.738083 18.692413 18.649821 18.610099
#> 5  20.810790 20.939553 21.059639 21.171631 21.276075 21.373480 21.464320
#> 6  22.032663 22.142498 22.244931 22.340460 22.429551 22.512637 22.590123
#> 7  24.060937 24.337147 24.594742 24.834975 25.059018 25.267961 25.462822
#> 8  20.080089 20.132241 20.180878 20.226237 20.268539 20.307989 20.344781
#> 9  15.960842 15.941012 15.922519 15.905272 15.889188 15.874188 15.860198
#> 10  9.402985  9.438662  9.471934  9.502963  9.531902  9.558889  9.584058
#> 11  9.549082  9.559857  9.569907  9.579279  9.588020  9.596171  9.603774
#> 12 26.687139 26.743530 26.796121 26.845167 26.890908 26.933566 26.973349
#> 13 29.036375 29.024449 29.013326 29.002953 28.993279 28.984257 28.975843
#> 14 28.797128 28.869895 28.937757 29.001046 29.060069 29.115114 29.166450
#> 15 18.398284 18.335175 18.276318 18.221429 18.170239 18.122498 18.077976
#> 16 19.097434 19.057167 19.019615 18.984593 18.951931 18.921471 18.893064
#> 17 16.874086 16.828867 16.786697 16.747368 16.710690 16.676484 16.644584
#> 18 25.244624 25.139830 25.042099 24.950954 24.865953 24.786680 24.712750
#> 19 18.668958 18.461040 18.267135 18.086298 17.917649 17.760367 17.613685
#> 20 16.716463 16.523554 16.343646 16.175863 16.019388 15.873460 15.737366
#> 21 23.216668 23.236267 23.254546 23.271592 23.287490 23.302316 23.316143
#>       0.6200    0.5782    0.5392    0.5029   0.4690   0.4374   0.4079   0.3804
#> 1  22.058232 22.041648 22.030357 22.010407 21.99175 21.97435 21.71054 21.46047
#> 2  25.595531 25.597004 25.547279 25.459969 25.37840 25.30232 25.36274 25.50143
#> 3  22.001762 22.013806 21.869705 21.616743 21.38041 21.16001 20.98078 21.02305
#> 4  18.573055 18.564373 18.452926 18.257151 18.07423 17.90363 17.87674 18.02291
#> 5  21.549038 21.568893 21.372129 21.026743 20.70407 20.40314 20.02278 19.94282
#> 6  22.662388 22.679324 22.705445 22.751441 22.79443 22.83453 22.82237 22.76928
#> 7  25.644550 25.687141 25.790910 25.973451 26.14404 26.30314 26.17615 25.88555
#> 8  20.379094 20.387135 20.502181 20.704271 20.89309 21.06919 20.98425 20.71898
#> 9  15.847152 15.844094 15.793502 15.704637 15.62161 15.54417 15.48587 15.50545
#> 10  9.607531  9.613032  9.719043  9.905248 10.07922 10.24148 10.50211 10.60163
#> 11  9.610863  9.612525  9.769859 10.046161 10.30431 10.54507 10.98165 11.17991
#> 12 27.010451 27.019147 27.023639 27.031593 27.03903 27.04597 27.14458 27.23761
#> 13 28.967996 28.966157 29.126213 29.407269 29.66986 29.91475 30.23799 30.31699
#> 14 29.214325 29.225546 29.225052 29.224269 29.22355 29.22288 29.21380 29.20529
#> 15 18.036454 18.026723 17.815987 17.445847 17.10001 16.77749 16.50408 16.55457
#> 16 18.866571 18.860362 18.750084 18.556382 18.37540 18.20661 18.15107 18.26633
#> 17 16.614833 16.607861 16.528980 16.390408 16.26093 16.14018 16.25061 16.48547
#> 18 24.643802 24.627643 24.703976 24.837902 24.96301 25.07969 25.44164 25.69030
#> 19 17.476889 17.444828 17.580370 17.818154 18.04028 18.24744 18.67580 18.89981
#> 20 15.610444 15.580698 15.578816 15.575291 15.57197 15.56886 14.71456 13.85064
#> 21 23.329038 23.332060 23.413546 23.556665 23.69038 23.81509 24.05988 24.18151
#>      0.3548   0.3309   0.3086   0.2878   0.2684   0.2503   0.2334   0.2177
#> 1  21.26326 21.26301 21.26283 21.26270 21.26260 21.26253 21.52225 22.73717
#> 2  25.61086 25.61090 25.61093 25.61095 25.61097 25.61098 26.41224 25.87168
#> 3  21.05651 21.05635 21.05623 21.05614 21.05608 21.05603 20.94192 20.93695
#> 4  18.13832 18.13826 18.13821 18.13817 18.13814 18.13812 17.55183 17.83972
#> 5  19.87993 19.87960 19.87936 19.87919 19.87906 19.87897 20.30796 19.93710
#> 6  22.72736 22.72739 22.72741 22.72743 22.72744 22.72745 22.12000 21.19832
#> 7  25.65617 25.65618 25.65618 25.65618 25.65619 25.65619 25.38207 25.06713
#> 8  20.50961 20.50959 20.50958 20.50956 20.50955 20.50955 19.47191 19.88029
#> 9  15.52095 15.52088 15.52083 15.52080 15.52077 15.52075 15.33510 15.81796
#> 10 10.68001 10.68026 10.68044 10.68058 10.68067 10.68075 11.30166 11.36653
#> 11 11.33613 11.33653 11.33682 11.33704 11.33719 11.33731 11.65243 11.91770
#> 12 27.31098 27.31107 27.31114 27.31120 27.31124 27.31126 28.01938 27.88272
#> 13 30.37915 30.37945 30.37967 30.37983 30.37995 30.38004 29.80320 31.48494
#> 14 29.19856 29.19858 29.19859 29.19860 29.19860 29.19860 29.66860 30.01790
#> 15 16.59462 16.59433 16.59411 16.59395 16.59383 16.59375 16.34076 16.22506
#> 16 18.35736 18.35727 18.35721 18.35716 18.35713 18.35710 17.82037 18.14993
#> 17 16.67080 16.67086 16.67090 16.67093 16.67095 16.67096 16.37450 16.25958
#> 18 25.88638 25.88666 25.88687 25.88702 25.88714 25.88722 25.67046 23.99372
#> 19 19.07641 19.07674 19.07698 19.07715 19.07728 19.07737 19.01969 17.49683
#> 20 13.16926 13.16851 13.16797 13.16756 13.16727 13.16705 13.54801 14.15955
#> 21 24.27735 24.27758 24.27774 24.27786 24.27795 24.27802 25.03566 25.05920
#>      0.2030   0.1893   0.1766   0.1647   0.1536   0.1432   0.1336   0.1246
#> 1  22.73702 22.73690 23.25493 23.25512 23.25526 22.77556 22.77516 22.77482
#> 2  25.87137 25.87112 25.55024 25.55042 25.55057 25.84573 25.84589 25.84601
#> 3  20.93682 20.93675 20.36836 20.36814 20.36796 20.43621 20.43648 20.43671
#> 4  17.83977 17.83986 18.01925 18.01895 18.01871 18.00100 18.00108 18.00115
#> 5  19.93672 19.93646 19.64143 19.64151 19.64158 19.72233 19.72245 19.72255
#> 6  21.19857 21.19879 20.90082 20.90063 20.90048 21.06543 21.06565 21.06584
#> 7  25.06737 25.06754 25.85947 25.85983 25.86012 25.65555 25.65507 25.65468
#> 8  19.88076 19.88113 19.11426 19.11382 19.11347 18.83545 18.83564 18.83581
#> 9  15.81796 15.81796 16.04417 16.04410 16.04405 15.91551 15.91542 15.91535
#> 10 11.36648 11.36639 11.53755 11.53787 11.53812 11.61627 11.61609 11.61594
#> 11 11.91783 11.91786 11.93766 11.93780 11.93793 12.03077 12.03068 12.03060
#> 12 27.88251 27.88233 27.89300 27.89328 27.89350 28.03177 28.03171 28.03165
#> 13 31.48534 31.48561 31.18370 31.18341 31.18319 30.83637 30.83628 30.83622
#> 14 30.01776 30.01762 30.24173 30.24198 30.24219 30.16636 30.16614 30.16596
#> 15 16.22486 16.22478 16.35640 16.35620 16.35603 16.39016 16.39029 16.39041
#> 16 18.14997 18.15004 18.43608 18.43584 18.43565 18.38164 18.38165 18.38167
#> 17 16.25960 16.25964 16.38915 16.38896 16.38881 16.54740 16.54752 16.54763
#> 18 23.99390 23.99406 24.73919 24.73927 24.73933 25.22625 25.22631 25.22635
#> 19 17.49702 17.49716 17.25671 17.25658 17.25648 17.77928 17.77967 17.78000
#> 20 14.15929 14.15910 13.93015 13.93031 13.93045 13.23699 13.23677 13.23660
#> 21 25.05907 25.05892 24.64576 24.64595 24.64611 24.80398 24.80403 24.80408
#>      0.1162   0.1083   0.1010   0.0942   0.0879   0.0820   0.0764   0.0713
#> 1  22.77455 22.77433 22.77414 22.77399 22.77387 22.77376 22.75462 22.75474
#> 2  25.84611 25.84618 25.84624 25.84629 25.84632 25.84635 26.02598 26.02583
#> 3  20.43689 20.43704 20.43717 20.43727 20.43736 20.43743 20.91459 20.91442
#> 4  18.00121 18.00127 18.00132 18.00136 18.00140 18.00143 18.33966 18.33954
#> 5  19.72262 19.72267 19.72272 19.72275 19.72278 19.72280 19.52403 19.52390
#> 6  21.06599 21.06612 21.06623 21.06632 21.06639 21.06645 21.73984 21.73998
#> 7  25.65435 25.65408 25.65385 25.65367 25.65352 25.65339 25.46876 25.46869
#> 8  18.83596 18.83609 18.83620 18.83629 18.83637 18.83644 18.19573 18.19626
#> 9  15.91529 15.91524 15.91521 15.91518 15.91516 15.91514 14.91591 14.91628
#> 10 11.61582 11.61571 11.61562 11.61554 11.61548 11.61543 11.96123 11.96108
#> 11 12.03053 12.03048 12.03043 12.03040 12.03037 12.03034 12.18722 12.18725
#> 12 28.03160 28.03155 28.03151 28.03147 28.03144 28.03141 27.63393 27.63387
#> 13 30.83618 30.83615 30.83613 30.83612 30.83612 30.83611 31.11021 31.11037
#> 14 30.16581 30.16568 30.16557 30.16548 30.16541 30.16535 29.83369 29.83356
#> 15 16.39050 16.39057 16.39064 16.39069 16.39074 16.39077 16.35234 16.35231
#> 16 18.38168 18.38170 18.38172 18.38173 18.38175 18.38176 17.95541 17.95551
#> 17 16.54772 16.54779 16.54786 16.54791 16.54795 16.54799 17.04314 17.04292
#> 18 25.22638 25.22640 25.22642 25.22643 25.22644 25.22645 25.79425 25.79424
#> 19 17.78026 17.78047 17.78065 17.78080 17.78092 17.78101 17.22690 17.22673
#> 20 13.23645 13.23633 13.23623 13.23614 13.23607 13.23601 13.37089 13.37082
#> 21 24.80411 24.80413 24.80414 24.80416 24.80416 24.80417 24.95169 24.95170
#>      0.0665   0.0620   0.0578   0.0539   0.0503   0.0469   0.0437   0.0408
#> 1  22.75484 22.75493 22.75502 22.75509 22.75516 22.75522 22.75527 22.75531
#> 2  26.02569 26.02555 26.02541 26.02528 26.02516 26.02503 26.02492 26.02481
#> 3  20.91426 20.91411 20.91397 20.91385 20.91373 20.91363 20.91353 20.91344
#> 4  18.33943 18.33933 18.33923 18.33915 18.33908 18.33901 18.33895 18.33890
#> 5  19.52377 19.52366 19.52357 19.52348 19.52341 19.52334 19.52329 19.52324
#> 6  21.74009 21.74017 21.74023 21.74026 21.74027 21.74027 21.74025 21.74022
#> 7  25.46865 25.46862 25.46860 25.46860 25.46861 25.46863 25.46865 25.46868
#> 8  18.19676 18.19722 18.19765 18.19805 18.19842 18.19875 18.19906 18.19935
#> 9  14.91662 14.91694 14.91723 14.91750 14.91775 14.91798 14.91820 14.91839
#> 10 11.96093 11.96080 11.96066 11.96054 11.96043 11.96032 11.96022 11.96012
#> 11 12.18727 12.18729 12.18729 12.18730 12.18729 12.18729 12.18728 12.18726
#> 12 27.63382 27.63377 27.63374 27.63371 27.63369 27.63367 27.63365 27.63364
#> 13 31.11053 31.11068 31.11082 31.11095 31.11107 31.11118 31.11129 31.11138
#> 14 29.83346 29.83337 29.83331 29.83326 29.83322 29.83320 29.83319 29.83318
#> 15 16.35229 16.35226 16.35224 16.35221 16.35219 16.35216 16.35214 16.35211
#> 16 17.95560 17.95570 17.95579 17.95588 17.95597 17.95605 17.95613 17.95620
#> 17 17.04272 17.04253 17.04235 17.04219 17.04205 17.04191 17.04179 17.04167
#> 18 25.79420 25.79414 25.79406 25.79397 25.79386 25.79375 25.79363 25.79351
#> 19 17.22660 17.22649 17.22641 17.22634 17.22630 17.22628 17.22626 17.22626
#> 20 13.37078 13.37075 13.37073 13.37073 13.37073 13.37075 13.37077 13.37080
#> 21 24.95170 24.95169 24.95167 24.95165 24.95162 24.95158 24.95154 24.95150
#>      0.0380   0.0355   0.0331   0.0309   0.0288   0.0268   0.0250   0.0233
#> 1  22.88900 22.88913 22.88925 22.88936 22.88947 22.88958 22.88968 22.88977
#> 2  25.95533 25.95532 25.95530 25.95529 25.95527 25.95526 25.95525 25.95524
#> 3  20.96935 20.96944 20.96952 20.96960 20.96968 20.96975 20.96982 20.96989
#> 4  18.40252 18.40262 18.40271 18.40280 18.40289 18.40297 18.40304 18.40312
#> 5  19.47139 19.47129 19.47120 19.47111 19.47103 19.47095 19.47087 19.47080
#> 6  21.77371 21.77389 21.77406 21.77423 21.77438 21.77453 21.77468 21.77481
#> 7  25.51912 25.51907 25.51903 25.51899 25.51895 25.51892 25.51888 25.51885
#> 8  18.18207 18.18197 18.18187 18.18177 18.18168 18.18159 18.18151 18.18143
#> 9  14.79553 14.79534 14.79516 14.79499 14.79483 14.79467 14.79452 14.79438
#> 10 11.98954 11.98961 11.98968 11.98975 11.98981 11.98986 11.98992 11.98997
#> 11 12.17874 12.17878 12.17881 12.17885 12.17888 12.17892 12.17895 12.17898
#> 12 27.53333 27.53318 27.53305 27.53291 27.53279 27.53267 27.53255 27.53244
#> 13 31.26339 31.26354 31.26368 31.26382 31.26395 31.26408 31.26420 31.26431
#> 14 29.81811 29.81801 29.81791 29.81782 29.81773 29.81765 29.81757 29.81749
#> 15 16.33134 16.33135 16.33135 16.33136 16.33136 16.33136 16.33137 16.33137
#> 16 17.91016 17.91008 17.91000 17.90993 17.90986 17.90980 17.90973 17.90967
#> 17 17.07517 17.07526 17.07535 17.07544 17.07553 17.07561 17.07568 17.07575
#> 18 25.71174 25.71186 25.71197 25.71207 25.71217 25.71227 25.71236 25.71245
#> 19 17.00574 17.00542 17.00512 17.00482 17.00455 17.00428 17.00403 17.00378
#> 20 13.60894 13.60905 13.60916 13.60926 13.60935 13.60945 13.60954 13.60962
#> 21 24.91577 24.91579 24.91580 24.91582 24.91583 24.91585 24.91586 24.91587
#>      0.0218   0.0203   0.0189   0.0177   0.0165   0.0154   0.0143   0.0134
#> 1  22.88986 22.88995 22.89003 22.89011 22.89019 22.89026 22.89033 22.89040
#> 2  25.95523 25.95522 25.95521 25.95520 25.95519 25.95518 25.95517 25.95516
#> 3  20.96995 20.97001 20.97007 20.97013 20.97018 20.97023 20.97028 20.97032
#> 4  18.40319 18.40325 18.40332 18.40338 18.40343 18.40349 18.40354 18.40359
#> 5  19.47073 19.47066 19.47060 19.47054 19.47048 19.47042 19.47037 19.47032
#> 6  21.77494 21.77507 21.77519 21.77530 21.77541 21.77551 21.77561 21.77571
#> 7  25.51882 25.51879 25.51876 25.51873 25.51871 25.51868 25.51866 25.51863
#> 8  18.18135 18.18128 18.18121 18.18115 18.18108 18.18103 18.18097 18.18091
#> 9  14.79424 14.79411 14.79399 14.79387 14.79376 14.79365 14.79354 14.79345
#> 10 11.99002 11.99007 11.99012 11.99016 11.99020 11.99024 11.99028 11.99032
#> 11 12.17901 12.17903 12.17906 12.17908 12.17911 12.17913 12.17915 12.17917
#> 12 27.53234 27.53224 27.53214 27.53205 27.53197 27.53188 27.53180 27.53173
#> 13 31.26442 31.26453 31.26463 31.26472 31.26481 31.26490 31.26498 31.26506
#> 14 29.81741 29.81734 29.81728 29.81721 29.81715 29.81709 29.81704 29.81698
#> 15 16.33137 16.33138 16.33138 16.33138 16.33139 16.33139 16.33139 16.33139
#> 16 17.90962 17.90956 17.90951 17.90946 17.90941 17.90937 17.90932 17.90928
#> 17 17.07582 17.07589 17.07595 17.07601 17.07607 17.07612 17.07618 17.07623
#> 18 25.71253 25.71261 25.71269 25.71276 25.71283 25.71290 25.71296 25.71302
#> 19 17.00355 17.00333 17.00312 17.00292 17.00273 17.00255 17.00237 17.00221
#> 20 13.60970 13.60978 13.60985 13.60992 13.60999 13.61005 13.61011 13.61017
#> 21 24.91588 24.91589 24.91590 24.91591 24.91592 24.91593 24.91594 24.91595
#>      0.0125   0.0116   0.0108   0.0101   0.0094   0.0088   0.0082   0.0076
#> 1  22.89046 22.89052 22.89058 22.89063 22.89068 22.89073 22.89078 22.78977
#> 2  25.95515 25.95515 25.95514 25.95513 25.95512 25.95512 25.95511 25.99428
#> 3  20.97037 20.97041 20.97045 20.97049 20.97052 20.97056 20.97059 20.97905
#> 4  18.40364 18.40369 18.40373 18.40377 18.40381 18.40385 18.40389 18.45734
#> 5  19.47027 19.47023 19.47018 19.47014 19.47010 19.47006 19.47002 19.43344
#> 6  21.77580 21.77588 21.77596 21.77604 21.77612 21.77619 21.77625 21.75300
#> 7  25.51861 25.51859 25.51857 25.51855 25.51853 25.51852 25.51850 25.55281
#> 8  18.18086 18.18081 18.18077 18.18072 18.18068 18.18064 18.18060 18.12927
#> 9  14.79335 14.79326 14.79318 14.79310 14.79302 14.79294 14.79287 14.85542
#> 10 11.99035 11.99038 11.99041 11.99044 11.99047 11.99050 11.99053 11.92784
#> 11 12.17919 12.17921 12.17923 12.17924 12.17926 12.17928 12.17929 12.21060
#> 12 27.53166 27.53159 27.53152 27.53146 27.53140 27.53134 27.53129 27.50355
#> 13 31.26514 31.26521 31.26528 31.26534 31.26541 31.26547 31.26552 31.27754
#> 14 29.81693 29.81688 29.81684 29.81679 29.81675 29.81671 29.81667 29.80669
#> 15 16.33140 16.33140 16.33140 16.33140 16.33140 16.33141 16.33141 16.30680
#> 16 17.90924 17.90920 17.90917 17.90913 17.90910 17.90907 17.90904 17.91532
#> 17 17.07627 17.07632 17.07636 17.07640 17.07644 17.07648 17.07652 17.07632
#> 18 25.71308 25.71313 25.71319 25.71324 25.71328 25.71333 25.71337 25.72522
#> 19 17.00205 17.00190 17.00175 17.00162 17.00148 17.00136 17.00124 16.96585
#> 20 13.61022 13.61028 13.61033 13.61037 13.61042 13.61046 13.61051 13.67053
#> 21 24.91596 24.91596 24.91597 24.91598 24.91599 24.91599 24.91600 24.96934
#>      0.0071   0.0066   0.0062   0.0058   0.0054   0.0050
#> 1  22.78958 22.78940 22.78922 22.78905 22.78888 22.78872
#> 2  25.99434 25.99440 25.99446 25.99451 25.99456 25.99461
#> 3  20.97894 20.97883 20.97872 20.97862 20.97851 20.97841
#> 4  18.45736 18.45738 18.45740 18.45741 18.45743 18.45744
#> 5  19.43330 19.43316 19.43302 19.43289 19.43277 19.43264
#> 6  21.75302 21.75304 21.75305 21.75307 21.75308 21.75309
#> 7  25.55281 25.55281 25.55280 25.55280 25.55280 25.55279
#> 8  18.12938 18.12950 18.12961 18.12971 18.12982 18.12993
#> 9  14.85579 14.85616 14.85652 14.85686 14.85720 14.85753
#> 10 11.92764 11.92745 11.92727 11.92709 11.92692 11.92675
#> 11 12.21072 12.21083 12.21094 12.21105 12.21115 12.21125
#> 12 27.50351 27.50348 27.50345 27.50342 27.50339 27.50336
#> 13 31.27758 31.27762 31.27766 31.27769 31.27772 31.27775
#> 14 29.80661 29.80653 29.80645 29.80637 29.80630 29.80623
#> 15 16.30674 16.30669 16.30664 16.30659 16.30655 16.30650
#> 16 17.91540 17.91548 17.91555 17.91563 17.91570 17.91577
#> 17 17.07620 17.07609 17.07598 17.07587 17.07576 17.07565
#> 18 25.72534 25.72545 25.72556 25.72566 25.72576 25.72586
#> 19 16.96575 16.96565 16.96556 16.96547 16.96539 16.96532
#> 20 13.67047 13.67041 13.67034 13.67028 13.67022 13.67016
#> 21 24.96950 24.96966 24.96981 24.96995 24.97009 24.97023
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 22.8 21.4 18.7 18.1 24.4 22.8 17.8 16.4 10.4 14.7 32.4 30.4 33.9 15.5
#> [16] 15.2 19.2 26.0 15.8 15.0 21.4
#> 
#> $X
#>               am       carb         cyl        disp       drat       gear
#>  [1,]  1.1547005  0.8815906 -0.05299989 -0.59400773  0.4721880  0.4082483
#>  [2,]  1.1547005 -0.9697497 -1.16599767 -1.01999869  0.3868528  0.4082483
#>  [3,] -0.8660254 -0.9697497 -0.05299989  0.20882139 -0.9273089 -1.0206207
#>  [4,] -0.8660254 -0.3526363  1.05999788  1.04441905 -0.8078397 -1.0206207
#>  [5,] -0.8660254 -0.9697497 -0.05299989 -0.06151902 -1.4734540 -1.0206207
#>  [6,] -0.8660254 -0.3526363 -1.16599767 -0.70296311  0.1137802  0.4082483
#>  [7,] -0.8660254 -0.3526363 -1.16599767 -0.75129670  0.5063220  0.4082483
#>  [8,] -0.8660254  0.8815906 -0.05299989 -0.53174751  0.5063220  0.4082483
#>  [9,] -0.8660254  0.2644772  1.05999788  0.35464138 -0.9443759 -1.0206207
#> [10,] -0.8660254  0.8815906  1.05999788  1.86363243 -1.0638452 -1.0206207
#> [11,] -0.8660254  0.8815906  1.05999788  1.69978976 -0.6713034 -1.0206207
#> [12,]  1.1547005 -0.9697497 -1.16599767 -1.26002821  0.7793946  0.4082483
#> [13,]  1.1547005 -0.3526363 -1.16599767 -1.28460461  2.2300926  0.4082483
#> [14,]  1.1547005 -0.9697497 -1.16599767 -1.32228843  1.0183331  0.4082483
#> [15,] -0.8660254 -0.3526363  1.05999788  0.70034943 -1.4734540 -1.0206207
#> [16,] -0.8660254 -0.3526363  1.05999788  0.58565955 -0.8078397 -1.0206207
#> [17,] -0.8660254 -0.3526363  1.05999788  1.37210440 -0.9273089 -1.0206207
#> [18,]  1.1547005 -0.3526363 -1.16599767 -0.91923544  1.3767408  1.8371173
#> [19,]  1.1547005  0.8815906  1.05999788  0.97068984  1.0183331  1.8371173
#> [20,]  1.1547005  3.3500445  1.05999788  0.56108315 -0.1422253  1.8371173
#> [21,]  1.1547005 -0.3526363 -1.16599767 -0.91350095  0.8305957  0.4082483
#>               hp       qsec         vs          wt
#>  [1,] -0.4298768 -0.8595171 -0.9534626 -0.63948434
#>  [2,] -0.6678086  0.2805298  1.0488088 -0.95511826
#>  [3,] -0.4298768  0.7206410  1.0488088 -0.01347707
#>  [4,]  0.4798625 -0.5625747 -0.9534626  0.22324837
#>  [5,] -0.4998567  1.1342394  1.0488088  0.24429063
#>  [6,] -1.1016842  1.0175835  1.0488088 -0.03977989
#>  [7,] -0.6398166  2.5553212  1.0488088 -0.08186442
#>  [8,] -0.2479289  0.4343036  1.0488088  0.22324837
#>  [9,]  0.5498424 -0.3610780 -0.9534626  0.88607961
#> [10,]  1.0397020 -0.1383712 -0.9534626  2.31064070
#> [11,]  1.2496418 -0.3504729 -0.9534626  2.22752377
#> [12,] -1.0457003  0.7365486  1.0488088 -1.08137183
#> [13,] -1.2416441  0.2328069  1.0488088 -1.69685798
#> [14,] -1.0596963  0.9645580  1.0488088 -1.46539310
#> [15,]  0.1299627 -0.6421128 -0.9534626  0.30741742
#> [16,]  0.1299627 -0.4141034 -0.9534626  0.21798781
#> [17,]  0.4798625 -0.5466670 -0.9534626  0.64935417
#> [18,] -0.6958006 -0.7322561 -0.9534626 -1.14449862
#> [19,]  1.7255054 -1.8988158 -0.9534626 -0.06082216
#> [20,]  2.7192206 -1.8457903 -0.9534626  0.36002307
#> [21,] -0.4438728  0.2752273  1.0488088 -0.47114625
#> attr(,"center")
#>  [1]   0.4285714   2.5714286   6.0952381 232.5095238   3.6233333   3.7142857
#>  [7] 140.7142857  18.0809524   0.4761905   3.2278095
#> attr(,"scale")
#>  [1]   0.4948717   1.6204476   1.7969488 122.0683177   0.5859249   0.6998542
#>  [7]  71.4490447   1.8858872   0.4994328   0.9504682
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 27
#> 
#> $lambda.min
#> [1] 0.8195794
#> 
#> $null.dev
#> [1] 35.83746
#> 
#> $Bias
#> [1] 5.714918
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 4.243949