Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 33.708513 32.094739 29.860792 27.057016 24.750721 22.868195 21.345927
#>   [8] 20.129237 19.171093 18.431074 17.874479 17.471551 17.196805 17.028442
#>  [15] 16.947848 16.809026 16.301799 15.959557 15.857508 15.781963 15.585975
#>  [22] 15.224608 14.263743 12.616705 11.420843 10.422207  9.409762  8.851326
#>  [29]  8.575276  8.505629  8.502271  8.534531  8.415810  8.266789 10.166045
#>  [36] 10.137893 11.793179 11.731919 12.708219 13.686496 13.187529 13.179008
#>  [43] 13.751441 15.833850 15.761189 15.761399 15.795234 15.655878 16.906439
#>  [50] 16.920521 24.864787 24.865408 25.084738 26.825983 28.652442 28.654459
#>  [57] 28.885922 28.444372 28.446148 35.274208 35.453200 35.456189 36.206323
#>  [64] 36.260000 36.318145 36.480322 36.460457 36.461860 36.479324 36.466870
#>  [71] 36.863655 36.864484 36.865286 36.892163 36.892865 36.893537 36.883899
#>  [78] 37.089334 37.091029 37.147455 37.147883 37.202220 37.202728 37.198238
#>  [85] 37.198732 37.199212 37.191491 37.191932 37.272117 37.272604 37.273082
#>  [92] 37.273554 37.274020 37.274480 37.274936 37.275386 37.275832 37.276274
#>  [99] 37.276713 37.277148
#> 
#> $cvse
#>   [1] 11.091872 11.022808 10.537333  9.596334  8.746677  7.979660  7.288348
#>   [8]  6.667283  6.112237  5.619975  5.188016  4.814384  4.497318  4.234966
#>  [15]  4.025078  3.842055  3.667088  3.525223  3.494162  3.454518  3.367595
#>  [22]  3.295914  3.186533  3.124335  3.094869  3.032346  2.914724  2.833895
#>  [29]  2.774348  2.771847  2.784517  2.794222  2.782985  2.774604  3.311404
#>  [36]  3.293431  3.148077  3.124202  3.372249  3.562921  3.364967  3.365096
#>  [43]  3.684385  4.286271  4.299954  4.299967  4.295504  4.318487  4.796295
#>  [50]  4.791328 10.253097 10.253831 10.234203 10.938366 11.313306 11.314410
#>  [57] 11.300192 11.276142 11.277037 11.798030 11.780258 11.781423 12.084835
#>  [64] 12.079385 12.099288 12.092426 12.093400 12.093228 12.093470 12.095341
#>  [71] 12.180253 12.180069 12.179895 12.178258 12.178107 12.177967 12.179484
#>  [78] 12.183126 12.183202 12.132312 12.131855 12.125092 12.124662 12.125003
#>  [85] 12.124612 12.124240 12.125125 12.124795 12.123100 12.122803 12.122522
#>  [92] 12.122258 12.122010 12.121777 12.121559 12.121356 12.121167 12.120991
#>  [99] 12.120830 12.120681
#> 
#> $fold
#>  [1]  8  1  9 10  5  9  2  1  3  7  1  4  6  5 10  6  8  3  7  4  2
#> 
#> $lambda
#>   [1] 4.737254618 4.417979511 4.120222479 3.842533274 3.583559391 3.342039482
#>   [7] 3.116797207 2.906735506 2.710831262 2.528130308 2.357742786 2.198838813
#>  [13] 2.050644437 1.912437865 1.783545954 1.663340926 1.551237314 1.446689111
#>  [19] 1.349187107 1.258256411 1.173454141 1.094367259 1.020610568 0.951824832
#>  [25] 0.887675024 0.827848698 0.772054467 0.720020580 0.671493602 0.626237181
#>  [31] 0.584030891 0.544669164 0.507960285 0.473725462 0.441797951 0.412022248
#>  [37] 0.384253327 0.358355939 0.334203948 0.311679721 0.290673551 0.271083126
#>  [43] 0.252813031 0.235774279 0.219883881 0.205064444 0.191243786 0.178354595
#>  [49] 0.166334093 0.155123731 0.144668911 0.134918711 0.125825641 0.117345414
#>  [55] 0.109436726 0.102061057 0.095182483 0.088767502 0.082784870 0.077205447
#>  [61] 0.072002058 0.067149360 0.062623718 0.058403089 0.054466916 0.050796028
#>  [67] 0.047372546 0.044179795 0.041202225 0.038425333 0.035835594 0.033420395
#>  [73] 0.031167972 0.029067355 0.027108313 0.025281303 0.023577428 0.021988388
#>  [79] 0.020506444 0.019124379 0.017835460 0.016633409 0.015512373 0.014466891
#>  [85] 0.013491871 0.012582564 0.011734541 0.010943673 0.010206106 0.009518248
#>  [91] 0.008876750 0.008278487 0.007720545 0.007200206 0.006714936 0.006262372
#>  [97] 0.005840309 0.005446692 0.005079603 0.004737255
#> 
#> $fit
#> $beta
#>               4.7373     4.4180     4.1202     3.8425     3.5836    3.3420
#> (Intercept) 19.31429 20.9339999 22.6008370 24.1553348 25.6050647 26.957088
#> am           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> cyl          0.00000 -0.2501029 -0.5074822 -0.7475149 -0.9713703 -1.180139
#> disp         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> wt           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#>                3.1168    2.9067    2.7108    2.5281    2.3577    2.1988
#> (Intercept) 28.217989 29.393909 30.490577 31.513333 32.467158 33.356699
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.374837 -1.556412 -1.725751 -1.883676 -2.030958 -2.168314
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                2.0506    1.9124   1.7835    1.6633    1.5512    1.4467
#> (Intercept) 34.186288 34.959965 35.68150 36.354404 36.881928 36.881928
#> am           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> cyl         -2.296412 -2.415877 -2.52729 -2.631195 -2.712651 -2.712651
#> disp         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#>                1.3492        1.2583    1.1735    1.0944    1.0206    0.9518
#> (Intercept) 36.881928 36.8837545935 37.333970 37.753860 38.145450 38.510636
#> am           0.000000  0.0000000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.0000000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.712651 -2.7124270921 -2.438942 -2.183571 -1.945411 -1.723534
#> disp         0.000000  0.0000000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.0000000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.0000000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.0000000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.0000000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.0000000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000 -0.0009611139 -0.652987 -1.261676 -1.829344 -2.358308
#>                0.8877    0.8278    0.7721    0.7200    0.6715    0.6262
#> (Intercept) 38.764465 38.764238 38.764116 38.764051 38.764016 38.763997
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.568585 -1.568478 -1.568421 -1.568390 -1.568374 -1.568365
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.727359 -2.727496 -2.727569 -2.727608 -2.727629 -2.727640
#>                  0.5840      0.5447     0.5080    0.4737    0.4418     0.4120
#> (Intercept) 38.77119873 38.78723733 38.8020942 38.815947 38.828865 38.8409131
#> am           0.00000000  0.00000000  0.0000000  0.000000  0.000000  0.0000000
#> carb        -0.02207558 -0.07297239 -0.1205119 -0.164850 -0.206200 -0.2447631
#> cyl         -1.55776511 -1.53306023 -1.5099267 -1.488349 -1.468226 -1.4494584
#> disp         0.00000000  0.00000000  0.0000000  0.000000  0.000000  0.0000000
#> drat         0.00000000  0.00000000  0.0000000  0.000000  0.000000  0.0000000
#> gear         0.00000000  0.00000000  0.0000000  0.000000  0.000000  0.0000000
#> hp           0.00000000  0.00000000  0.0000000  0.000000  0.000000  0.0000000
#> qsec         0.00000000  0.00000000  0.0000000  0.000000  0.000000  0.0000000
#> vs           0.00000000  0.00000000  0.0000000  0.000000  0.000000  0.0000000
#> wt          -2.73015601 -2.73629598 -2.7421054 -2.747526 -2.752581 -2.7572961
#>                 0.3843     0.3584    0.3342    0.3117    0.2907    0.2711
#> (Intercept) 38.8521491 38.8626278 34.735326 34.734848 34.734496 34.734237
#> am           0.0000000  0.0000000  3.491296  3.491901  3.492349  3.492680
#> carb        -0.2807272 -0.3142674 -1.098806 -1.098886 -1.098945 -1.098988
#> cyl         -1.4319560 -1.4156333 -1.299096 -1.299224 -1.299319 -1.299390
#> disp         0.0000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.0000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.0000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.0000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.0000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.0000000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.7616931 -2.7657936 -1.464435 -1.464049 -1.463762 -1.463549
#>                0.2528    0.2358    0.2199    0.2051    0.1912    0.1784
#> (Intercept) 34.734046 34.733905 34.733802 34.733725 34.733670 34.733628
#> am           3.492925  3.493106  3.493240  3.493338  3.493411  3.493464
#> carb        -1.099020 -1.099043 -1.099060 -1.099073 -1.099082 -1.099089
#> cyl         -1.299443 -1.299482 -1.299512 -1.299533 -1.299549 -1.299561
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -1.463391 -1.463275 -1.463188 -1.463124 -1.463077 -1.463042
#>                0.1663    0.1551    0.1447    0.1349    0.1258    0.1173
#> (Intercept) 34.733598 34.733576 32.407586 32.405960 32.404670 32.403647
#> am           3.493504  3.493533  3.866126  3.866959  3.867632  3.868173
#> carb        -1.099094 -1.099097 -1.075883 -1.076019 -1.076127 -1.076214
#> cyl         -1.299570 -1.299577 -1.038332 -1.038226 -1.038148 -1.038090
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  1.211728  1.212134  1.212453  1.212704
#> wt          -1.463016 -1.462997 -1.440918 -1.440648 -1.440422 -1.440236
#>                  0.1094      0.1021      0.0952      0.0888      0.0828
#> (Intercept) 24.36944906 24.36831537 24.36737732 24.36660611 24.36597672
#> am           3.07067830  3.07084668  3.07100450  3.07115155  3.07128785
#> carb        -0.69049694 -0.69045080 -0.69041435 -0.69038600 -0.69036437
#> cyl         -0.38561068 -0.38535223 -0.38513288 -0.38494739 -0.38479116
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         1.33455355  1.33450739  1.33445896  1.33440957  1.33436026
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp          -0.01586312 -0.01586548 -0.01586743 -0.01586904 -0.01587036
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           1.85825097  1.85877862  1.85923099  1.85961766  1.85994710
#> wt          -1.29984078 -1.29994838 -1.30003900 -1.30011497 -1.30017835
#>                  0.0772      0.0720      0.0671      0.0626        0.0584
#> (Intercept) 24.36546742 24.36505943 24.36473654 24.36448481  2.436473e+01
#> am           3.07141357  3.07152904  3.07163467  3.07173093  3.071871e+00
#> carb        -0.69034831 -0.69033683 -0.69032906 -0.69032431 -6.903136e-01
#> cyl         -0.38466017 -0.38455085 -0.38446011 -0.38438525 -3.842665e-01
#> disp         0.00000000  0.00000000  0.00000000  0.00000000 -1.159771e-06
#> drat         1.33431184  1.33426489  1.33421988  1.33417710  1.333957e+00
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.000000e+00
#> hp          -0.01587142 -0.01587228 -0.01587296 -0.01587349 -1.587344e-02
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.000000e+00
#> vs           1.86022682  1.86046345  1.86066284  1.86083012  1.861068e+00
#> wt          -1.30023093 -1.30027428 -1.30030979 -1.30033863 -1.300370e+00
#>                  0.0545       0.0508       0.0474       0.0442       0.0412
#> (Intercept) 21.03968150 21.037958052 21.036537058 21.035396800 21.034516571
#> am           3.24714978  3.246983673  3.246825506  3.246675074  3.246532172
#> carb        -0.93653444 -0.936310999 -0.936094485 -0.935884961 -0.935682465
#> cyl         -0.15968360 -0.159396722 -0.159140140 -0.158912027 -0.158710625
#> disp        -0.00906435 -0.009060859 -0.009057277 -0.009053623 -0.009049915
#> drat         1.70606519  1.706234387  1.706366518  1.706464372  1.706530602
#> gear         0.00000000  0.000000000  0.000000000  0.000000000  0.000000000
#> hp          -0.01038729 -0.010394745 -0.010401826 -0.010408541 -0.010414900
#> qsec         0.00000000  0.000000000  0.000000000  0.000000000  0.000000000
#> vs           1.71369899  1.714392878  1.715042587  1.715649940  1.716216734
#> wt          -0.50497598 -0.505350023 -0.505731505 -0.506118619 -0.506509699
#>                   0.0384       0.0358      0.0334      0.0312      0.0291
#> (Intercept) 21.033876654 21.033458298 24.55326188 24.55184952 24.55050870
#> am           3.246396584  3.246268096  2.91411861  2.91379452  2.91348953
#> carb        -0.935487009 -0.935298583 -0.96687134 -0.96675471 -0.96663944
#> cyl         -0.158534247 -0.158381277 -0.19095434 -0.19065727 -0.19037688
#> disp        -0.009046168 -0.009042398 -0.01058926 -0.01058977 -0.01059011
#> drat         1.706567731  1.706578152  1.78191092  1.78231039  1.78268086
#> gear         0.000000000  0.000000000  0.00000000  0.00000000  0.00000000
#> hp          -0.010420916 -0.010426599 -0.01127109 -0.01127674 -0.01128218
#> qsec         0.000000000  0.000000000 -0.21326502 -0.21335298 -0.21343327
#> vs           1.716744734  1.717235672  2.11604372  2.11669864  2.11731898
#> wt          -0.506903211 -0.507297746 -0.25892918 -0.25886403 -0.25881911
#>                  0.0271      0.0253      0.0236       0.0220       0.0205
#> (Intercept) 24.54923719 24.54803275 24.54689312 25.727908784 25.730165871
#> am           2.91320289  2.91293382  2.91268160  3.040950629  3.041118680
#> carb        -0.96652570 -0.96641362 -0.96630333 -0.940772501 -0.940927511
#> cyl         -0.19011255 -0.18986368 -0.18962968 -0.283370533 -0.283608067
#> disp        -0.01059028 -0.01059029 -0.01059016 -0.010199248 -0.010203509
#> drat         1.78302370  1.78334021  1.78363167  1.788703472  1.788868754
#> gear         0.00000000  0.00000000  0.00000000 -0.230846681 -0.231182318
#> hp          -0.01128740 -0.01129242 -0.01129724 -0.009810298 -0.009802975
#> qsec        -0.21350626 -0.21357237 -0.21363195 -0.208733156 -0.208790370
#> vs           2.11790595  2.11846072  2.11898445  2.027291469  2.026849072
#> wt          -0.25879302 -0.25878437 -0.25879188 -0.335626968 -0.335211037
#>                   0.0191       0.0178       0.0166       0.0155       0.0145
#> (Intercept) 25.732267073 25.734218337 25.736025482 25.737694200 25.739230054
#> am           3.041267412  3.041397697  3.041510386  3.041606314  3.041686293
#> carb        -0.941079530 -0.941228491 -0.941374339 -0.941517026 -0.941656512
#> cyl         -0.283826380 -0.284026274 -0.284208531 -0.284373915 -0.284523166
#> disp        -0.010207765 -0.010212012 -0.010216242 -0.010220450 -0.010224631
#> drat         1.789042278  1.789223317  1.789411169  1.789605161  1.789804648
#> gear        -0.231496420 -0.231789781 -0.232063179 -0.232317376 -0.232553116
#> hp          -0.009796021 -0.009789427 -0.009783181 -0.009777272 -0.009771689
#> qsec        -0.208849952 -0.208911669 -0.208975299 -0.209040631 -0.209107462
#> vs           2.026439643  2.026061850  2.025714391  2.025396000  2.025105442
#> wt          -0.334792026 -0.334370772 -0.333948066 -0.333524654 -0.333101236
#>                   0.0135       0.0126       0.0117       0.0109       0.0102
#> (Intercept) 25.740638480 25.741924788 25.743094162 25.744151656 25.745102203
#> am           3.041751119  3.041801563  3.041838379  3.041862298  3.041874029
#> carb        -0.941792767 -0.941925767 -0.942055493 -0.942181937 -0.942305093
#> cyl         -0.284657008 -0.284776141 -0.284881251 -0.284973001 -0.285052036
#> disp        -0.010228780 -0.010232892 -0.010236964 -0.010240991 -0.010244971
#> drat         1.790009011  1.790217662  1.790430035  1.790645592  1.790863822
#> gear        -0.232771127 -0.232972119 -0.233156788 -0.233325810 -0.233479848
#> hp          -0.009766423 -0.009761462 -0.009756796 -0.009752415 -0.009748309
#> qsec        -0.209175601 -0.209244863 -0.209315076 -0.209386075 -0.209457702
#> vs           2.024841515  2.024603048  2.024388903  2.024197970  2.024029175
#> wt          -0.332678473 -0.332256984 -0.331837350 -0.331420112 -0.331005779
#>                   0.0095       0.0089       0.0083      0.0077       0.0072
#> (Intercept) 25.745950604 25.746701540 25.747359564 25.74792910 25.748414467
#> am           3.041874261  3.041863660  3.041842872  3.04181252  3.041773212
#> carb        -0.942424963 -0.942541552 -0.942654871 -0.94276494 -0.942871770
#> cyl         -0.285118983 -0.285174449 -0.285219024 -0.28525328 -0.285277771
#> disp        -0.010248898 -0.010252771 -0.010256586 -0.01026034 -0.010264035
#> drat         1.791084237  1.791306373  1.791529791  1.79175408  1.791978833
#> gear        -0.233619546 -0.233745534 -0.233858423 -0.23395881 -0.234047274
#> hp          -0.009744468 -0.009740882 -0.009737543 -0.00973444 -0.009731565
#> qsec        -0.209529810 -0.209602258 -0.209674913 -0.20974765 -0.209820355
#> vs           2.023881469  2.023753837  2.023645290  2.02355487  2.023481656
#> wt          -0.330594821 -0.330187678 -0.329784757 -0.32938643 -0.328993055
#>                   0.0067       0.0063       0.0058       0.0054       0.0051
#> (Intercept) 25.748819835 25.749149269 25.749406708 25.749595971 25.749720760
#> am           3.041725525  3.041670024  3.041607248  3.041537719  3.041461940
#> carb        -0.942975393 -0.943075834 -0.943173124 -0.943267297 -0.943358390
#> cyl         -0.285293031 -0.285299581 -0.285297921 -0.285288535 -0.285271893
#> disp        -0.010267665 -0.010271228 -0.010274723 -0.010278150 -0.010281506
#> drat         1.792203691  1.792428300  1.792652329  1.792875468  1.793097427
#> gear        -0.234124382 -0.234190680 -0.234246702 -0.234292965 -0.234329970
#> hp          -0.009728909 -0.009726462 -0.009724217 -0.009722165 -0.009720297
#> qsec        -0.209892913 -0.209965221 -0.210037183 -0.210108706 -0.210179705
#> vs           2.023424739  2.023383251  2.023356347  2.023343212  2.023343055
#> wt          -0.328604940 -0.328222381 -0.327845642 -0.327474966 -0.327110569
#>                   0.0047
#> (Intercept) 25.749784659
#> am           3.041380390
#> carb        -0.943446443
#> cyl         -0.285248446
#> disp        -0.010284790
#> drat         1.793317934
#> gear        -0.234358204
#> hp          -0.009718607
#> qsec        -0.210250103
#> vs           2.023355114
#> wt          -0.326752647
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   2  26  26  26  25  13   1   1   1   1   1   6   6   6   6   6   6
#>  [37]   6   6  40   1   1   1   1   1   1   1   1   1   1   1  41   1   1   1
#>  [55] 104   1   1   1   1   1   1   1   1   2  96   1   1   1   1   1   1  83
#>  [73]   1   1   1   1   1  66   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.737254618 4.417979511 4.120222479 3.842533274 3.583559391 3.342039482
#>   [7] 3.116797207 2.906735506 2.710831262 2.528130308 2.357742786 2.198838813
#>  [13] 2.050644437 1.912437865 1.783545954 1.663340926 1.551237314 1.446689111
#>  [19] 1.349187107 1.258256411 1.173454141 1.094367259 1.020610568 0.951824832
#>  [25] 0.887675024 0.827848698 0.772054467 0.720020580 0.671493602 0.626237181
#>  [31] 0.584030891 0.544669164 0.507960285 0.473725462 0.441797951 0.412022248
#>  [37] 0.384253327 0.358355939 0.334203948 0.311679721 0.290673551 0.271083126
#>  [43] 0.252813031 0.235774279 0.219883881 0.205064444 0.191243786 0.178354595
#>  [49] 0.166334093 0.155123731 0.144668911 0.134918711 0.125825641 0.117345414
#>  [55] 0.109436726 0.102061057 0.095182483 0.088767502 0.082784870 0.077205447
#>  [61] 0.072002058 0.067149360 0.062623718 0.058403089 0.054466916 0.050796028
#>  [67] 0.047372546 0.044179795 0.041202225 0.038425333 0.035835594 0.033420395
#>  [73] 0.031167972 0.029067355 0.027108313 0.025281303 0.023577428 0.021988388
#>  [79] 0.020506444 0.019124379 0.017835460 0.016633409 0.015512373 0.014466891
#>  [85] 0.013491871 0.012582564 0.011734541 0.010943673 0.010206106 0.009518248
#>  [91] 0.008876750 0.008278487 0.007720545 0.007200206 0.006714936 0.006262372
#>  [97] 0.005840309 0.005446692 0.005079603 0.004737255
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 19
#> 
#> $loss
#>   [1] 646.90571 565.05461 489.08205 425.78011 373.31140 330.09044 294.75004
#>   [8] 266.11204 243.16196 225.02699 210.95690 200.30749 192.52616 187.13955
#>  [15] 183.74264 181.98947 181.56988 181.56988 181.56988 181.51937 151.35405
#>  [22] 130.58786 117.65564 111.19455 109.88010 109.88009 109.88009 109.88009
#>  [29] 109.88009 109.88009 108.96887 107.04192 105.46121 104.17772 103.14666
#>  [36] 102.32941 101.69275 101.20817  84.86662  84.86661  84.86661  84.86661
#>  [43]  84.86661  84.86661  84.86661  84.86661  84.86661  84.86661  84.86661
#>  [50]  84.86661  83.14657  83.14657  83.14656  83.14656  79.04281  79.04281
#>  [57]  79.04280  79.04280  79.04280  79.04279  79.04279  79.04279  79.04279
#>  [64]  79.04244  77.67779  77.67777  77.67775  77.67773  77.67771  77.67769
#>  [71]  77.67767  77.38803  77.38802  77.38800  77.38799  77.38798  77.38797
#>  [78]  77.30128  77.30126  77.30124  77.30122  77.30120  77.30118  77.30116
#>  [85]  77.30115  77.30113  77.30112  77.30110  77.30109  77.30108  77.30107
#>  [92]  77.30106  77.30104  77.30103  77.30102  77.30102  77.30101  77.30100
#>  [99]  77.30099  77.30098
#> 
#> $linear.predictors
#>      4.7373   4.4180   4.1202   3.8425   3.5836   3.3420   3.1168   2.9067
#> 1  19.31429 19.43338 19.55594 19.67025 19.77684 19.87626 19.96897 20.05543
#> 2  19.31429 19.43338 19.55594 19.67025 19.77684 19.87626 19.96897 20.05543
#> 3  19.31429 18.93318 18.54098 18.17522 17.83410 17.51598 17.21930 16.94261
#> 4  19.31429 19.43338 19.55594 19.67025 19.77684 19.87626 19.96897 20.05543
#> 5  19.31429 19.93359 20.57091 21.16528 21.71958 22.23653 22.71864 23.16826
#> 6  19.31429 19.93359 20.57091 21.16528 21.71958 22.23653 22.71864 23.16826
#> 7  19.31429 18.93318 18.54098 18.17522 17.83410 17.51598 17.21930 16.94261
#> 8  19.31429 18.93318 18.54098 18.17522 17.83410 17.51598 17.21930 16.94261
#> 9  19.31429 18.93318 18.54098 18.17522 17.83410 17.51598 17.21930 16.94261
#> 10 19.31429 18.93318 18.54098 18.17522 17.83410 17.51598 17.21930 16.94261
#> 11 19.31429 18.93318 18.54098 18.17522 17.83410 17.51598 17.21930 16.94261
#> 12 19.31429 18.93318 18.54098 18.17522 17.83410 17.51598 17.21930 16.94261
#> 13 19.31429 19.93359 20.57091 21.16528 21.71958 22.23653 22.71864 23.16826
#> 14 19.31429 19.93359 20.57091 21.16528 21.71958 22.23653 22.71864 23.16826
#> 15 19.31429 19.93359 20.57091 21.16528 21.71958 22.23653 22.71864 23.16826
#> 16 19.31429 18.93318 18.54098 18.17522 17.83410 17.51598 17.21930 16.94261
#> 17 19.31429 18.93318 18.54098 18.17522 17.83410 17.51598 17.21930 16.94261
#> 18 19.31429 19.93359 20.57091 21.16528 21.71958 22.23653 22.71864 23.16826
#> 19 19.31429 18.93318 18.54098 18.17522 17.83410 17.51598 17.21930 16.94261
#> 20 19.31429 19.43338 19.55594 19.67025 19.77684 19.87626 19.96897 20.05543
#> 21 19.31429 18.93318 18.54098 18.17522 17.83410 17.51598 17.21930 16.94261
#>      2.7108   2.5281   2.3577   2.1988   2.0506   1.9124   1.7835   1.6633
#> 1  20.13607 20.21127 20.28141 20.34682 20.40782 20.46470 20.51776 20.56724
#> 2  20.13607 20.21127 20.28141 20.34682 20.40782 20.46470 20.51776 20.56724
#> 3  16.68457 16.44392 16.21949 16.01019 15.81499 15.63295 15.46318 15.30485
#> 4  20.13607 20.21127 20.28141 20.34682 20.40782 20.46470 20.51776 20.56724
#> 5  23.58757 23.97863 24.34333 24.68344 25.00064 25.29646 25.57234 25.82963
#> 6  23.58757 23.97863 24.34333 24.68344 25.00064 25.29646 25.57234 25.82963
#> 7  16.68457 16.44392 16.21949 16.01019 15.81499 15.63295 15.46318 15.30485
#> 8  16.68457 16.44392 16.21949 16.01019 15.81499 15.63295 15.46318 15.30485
#> 9  16.68457 16.44392 16.21949 16.01019 15.81499 15.63295 15.46318 15.30485
#> 10 16.68457 16.44392 16.21949 16.01019 15.81499 15.63295 15.46318 15.30485
#> 11 16.68457 16.44392 16.21949 16.01019 15.81499 15.63295 15.46318 15.30485
#> 12 16.68457 16.44392 16.21949 16.01019 15.81499 15.63295 15.46318 15.30485
#> 13 23.58757 23.97863 24.34333 24.68344 25.00064 25.29646 25.57234 25.82963
#> 14 23.58757 23.97863 24.34333 24.68344 25.00064 25.29646 25.57234 25.82963
#> 15 23.58757 23.97863 24.34333 24.68344 25.00064 25.29646 25.57234 25.82963
#> 16 16.68457 16.44392 16.21949 16.01019 15.81499 15.63295 15.46318 15.30485
#> 17 16.68457 16.44392 16.21949 16.01019 15.81499 15.63295 15.46318 15.30485
#> 18 23.58757 23.97863 24.34333 24.68344 25.00064 25.29646 25.57234 25.82963
#> 19 16.68457 16.44392 16.21949 16.01019 15.81499 15.63295 15.46318 15.30485
#> 20 20.13607 20.21127 20.28141 20.34682 20.40782 20.46470 20.51776 20.56724
#> 21 16.68457 16.44392 16.21949 16.01019 15.81499 15.63295 15.46318 15.30485
#>      1.5512   1.4467   1.3492   1.2583   1.1735   1.0944   1.0206   0.9518
#> 1  20.60602 20.60602 20.60602 20.60667 20.98949 21.34684 21.68010 21.99067
#> 2  20.60602 20.60602 20.60602 20.60643 20.82298 21.02511 21.21362 21.38930
#> 3  15.18072 15.18072 15.18072 15.18103 15.57616 15.94512 16.28922 16.60979
#> 4  20.60602 20.60602 20.60602 20.60587 20.44098 20.28703 20.14346 20.00969
#> 5  26.03133 26.03133 26.03133 26.03098 25.49517 24.99483 24.52820 24.09350
#> 6  26.03133 26.03133 26.03133 26.03102 25.52129 25.04529 24.60137 24.18783
#> 7  15.18072 15.18072 15.18072 15.18043 15.16478 15.15027 15.13673 15.12405
#> 8  15.18072 15.18072 15.18072 15.18075 15.38680 15.57924 15.75871 15.92588
#> 9  15.18072 15.18072 15.18072 15.18070 15.35415 15.51615 15.66724 15.80796
#> 10 15.18072 15.18072 15.18072 15.17929 14.39425 13.66149 12.97811 12.34125
#> 11 15.18072 15.18072 15.18072 15.17912 14.28064 13.44196 12.65980 11.93090
#> 12 15.18072 15.18072 15.18072 15.17920 14.33222 13.54163 12.80432 12.11721
#> 13 26.03133 26.03133 26.03133 26.03193 26.14163 26.24389 26.33925 26.42822
#> 14 26.03133 26.03133 26.03133 26.03249 26.52363 26.98197 27.40942 27.80783
#> 15 26.03133 26.03133 26.03133 26.03168 25.96859 25.90954 25.85447 25.80327
#> 16 15.18072 15.18072 15.18072 15.18104 15.57943 15.95143 16.29837 16.62158
#> 17 15.18072 15.18072 15.18072 15.18064 15.31170 15.43414 15.54834 15.65467
#> 18 26.03133 26.03133 26.03133 26.03199 26.18081 26.31959 26.44901 26.56972
#> 19 15.18072 15.18072 15.18072 15.18129 15.75247 16.28578 16.78314 17.24653
#> 20 20.60602 20.60602 20.60602 20.60653 20.89155 21.15759 21.40570 21.63692
#> 21 15.18072 15.18072 15.18072 15.18091 15.49127 15.78111 16.05141 16.30321
#>      0.8877   0.8278   0.7721   0.7200   0.6715   0.6262   0.5840   0.5447
#> 1  22.20727 22.20733 22.20736 22.20738 22.20738 22.20739 22.18330 22.12789
#> 2  21.51180 21.51182 21.51183 21.51184 21.51184 21.51184 21.48711 21.43014
#> 3  16.83367 16.83383 16.83391 16.83396 16.83398 16.83399 16.87319 16.96395
#> 4  19.91629 19.91623 19.91620 19.91619 19.91618 19.91617 19.95619 20.04832
#> 5  23.78985 23.78961 23.78949 23.78942 23.78938 23.78936 23.78679 23.78027
#> 6  23.89894 23.89871 23.89859 23.89852 23.89849 23.89847 23.89600 23.88972
#> 7  15.11543 15.11550 15.11554 15.11556 15.11557 15.11558 15.13112 15.16711
#> 8  16.04273 16.04285 16.04292 16.04295 16.04297 16.04298 16.05937 16.09745
#> 9  15.90637 15.90648 15.90654 15.90657 15.90659 15.90660 15.92286 15.96064
#> 10 11.89715 11.89706 11.89701 11.89699 11.89697 11.89697 11.88746 11.86531
#> 11 11.42259 11.42248 11.42241 11.42238 11.42237 11.42236 11.41241 11.38920
#> 12 11.63805 11.63795 11.63789 11.63786 11.63785 11.63784 11.62809 11.60536
#> 13 26.48993 26.48983 26.48978 26.48975 26.48974 26.48973 26.51172 26.56217
#> 14 28.08544 28.08542 28.08541 28.08540 28.08540 28.08540 28.08679 28.08993
#> 15 25.76718 25.76705 25.76698 25.76694 25.76692 25.76690 25.78823 25.83705
#> 16 16.84730 16.84746 16.84755 16.84759 16.84762 16.84763 16.88684 16.97763
#> 17 15.72909 15.72919 15.72925 15.72928 15.72929 15.72930 15.76748 15.85575
#> 18 26.65358 26.65348 26.65343 26.65341 26.65339 26.65339 26.65345 26.65338
#> 19 17.57006 17.57025 17.57035 17.57041 17.57044 17.57046 17.56618 17.55681
#> 20 21.79817 21.79821 21.79822 21.79823 21.79824 21.79824 21.72962 21.57150
#> 21 16.47911 16.47925 16.47933 16.47937 16.47939 16.47940 16.38582 16.17040
#>      0.5080   0.4737   0.4418   0.4120   0.3843   0.3584   0.3342   0.3117
#> 1  22.07617 22.02793 21.98295 21.94099 21.90187 21.86538 22.20000 22.20006
#> 2  21.37693 21.32731 21.28104 21.23788 21.19764 21.16010 21.82657 21.82672
#> 3  17.04881 17.12796 17.20178 17.27062 17.33482 17.39470 17.10729 17.10696
#> 4  20.13434 20.21456 20.28938 20.35915 20.42423 20.48491 20.77500 20.77501
#> 5  23.77405 23.76824 23.76283 23.75778 23.75307 23.74868 22.66978 22.66987
#> 6  23.88373 23.87814 23.87293 23.86807 23.86354 23.85931 22.72836 22.72843
#> 7  15.20078 15.23217 15.26145 15.28876 15.31423 15.33798 15.08589 15.08572
#> 8  16.13309 16.16633 16.19733 16.22624 16.25320 16.27835 15.58380 15.58350
#> 9  15.99599 16.02896 16.05970 16.08838 16.11512 16.14006 15.51057 15.51030
#> 10 11.84458 11.82524 11.80721 11.79039 11.77470 11.76008 12.25905 12.25926
#> 11 11.36745 11.34717 11.32826 11.31062 11.29417 11.27883 12.00424 12.00451
#> 12 11.58408 11.56423 11.54571 11.52845 11.51234 11.49732 12.11993 12.12017
#> 13 26.60924 26.65314 26.69408 26.73226 26.76787 26.80108 28.70967 28.71006
#> 14 28.09286 28.09560 28.09814 28.10052 28.10274 28.10480 28.46756 28.46764
#> 15 25.88259 25.92505 25.96465 26.00158 26.03602 26.06815 24.83030 24.83019
#> 16 17.06252 17.14170 17.21554 17.28441 17.34863 17.40853 17.11461 17.11428
#> 17 15.93826 16.01522 16.08698 16.15392 16.21634 16.27455 16.51419 16.51402
#> 18 26.65326 26.65314 26.65304 26.65294 26.65285 26.65276 27.69873 27.69902
#> 19 17.54816 17.54010 17.53258 17.52556 17.51902 17.51293 18.79637 18.79638
#> 20 21.42383 21.28610 21.15766 21.03787 20.92616 20.82198 19.78272 19.78268
#> 21 15.96927 15.78169 15.60674 15.44359 15.29144 15.14954 13.81537 13.81522
#>      0.2907   0.2711   0.2528   0.2358   0.2199   0.2051   0.1912   0.1784
#> 1  22.20010 22.20013 22.20015 22.20017 22.20018 22.20019 22.20019 22.20020
#> 2  21.82684 21.82692 21.82699 21.82703 21.82707 21.82709 21.82711 21.82712
#> 3  17.10671 17.10653 17.10640 17.10630 17.10622 17.10617 17.10613 17.10609
#> 4  20.77502 20.77503 20.77503 20.77504 20.77504 20.77504 20.77504 20.77505
#> 5  22.66993 22.66998 22.67002 22.67004 22.67007 22.67008 22.67009 22.67010
#> 6  22.72848 22.72852 22.72855 22.72858 22.72859 22.72861 22.72862 22.72862
#> 7  15.08560 15.08551 15.08544 15.08539 15.08535 15.08533 15.08530 15.08529
#> 8  15.58328 15.58312 15.58299 15.58290 15.58284 15.58279 15.58275 15.58272
#> 9  15.51009 15.50994 15.50982 15.50974 15.50968 15.50963 15.50960 15.50957
#> 10 12.25942 12.25953 12.25962 12.25968 12.25973 12.25977 12.25979 12.25981
#> 11 12.00472 12.00488 12.00499 12.00507 12.00514 12.00518 12.00522 12.00524
#> 12 12.12036 12.12050 12.12060 12.12067 12.12073 12.12077 12.12080 12.12082
#> 13 28.71035 28.71056 28.71072 28.71084 28.71092 28.71098 28.71103 28.71107
#> 14 28.46770 28.46775 28.46778 28.46781 28.46783 28.46784 28.46785 28.46786
#> 15 24.83010 24.83004 24.82999 24.82996 24.82994 24.82992 24.82991 24.82990
#> 16 17.11403 17.11385 17.11371 17.11361 17.11354 17.11348 17.11344 17.11341
#> 17 16.51389 16.51380 16.51372 16.51367 16.51363 16.51360 16.51358 16.51356
#> 18 27.69923 27.69939 27.69950 27.69959 27.69965 27.69970 27.69973 27.69976
#> 19 18.79639 18.79640 18.79640 18.79640 18.79640 18.79640 18.79640 18.79640
#> 20 19.78264 19.78262 19.78260 19.78259 19.78258 19.78257 19.78257 19.78257
#> 21 13.81511 13.81502 13.81496 13.81492 13.81489 13.81486 13.81484 13.81483
#>      0.1663   0.1551   0.1447   0.1349   0.1258   0.1173   0.1094   0.1021
#> 1  22.20020 22.20021 21.96498 21.96499 21.96500 21.96501 22.41871 22.41876
#> 2  21.82713 21.82714 21.59755 21.59763 21.59769 21.59775 22.08725 22.08727
#> 3  17.10607 17.10606 16.99241 16.99229 16.99218 16.99209 16.85992 16.86001
#> 4  20.77505 20.77505 21.32786 21.32808 21.32825 21.32838 20.74383 20.74407
#> 5  22.67011 22.67011 22.71769 22.71749 22.71733 22.71721 23.09876 23.09862
#> 6  22.72863 22.72863 22.77533 22.77511 22.77495 22.77482 22.93422 22.93399
#> 7  15.08528 15.08527 15.00875 15.00866 15.00859 15.00852 15.16444 15.16451
#> 8  15.58270 15.58269 15.49866 15.49848 15.49833 15.49820 15.60638 15.60649
#> 9  15.50955 15.50954 15.42661 15.42645 15.42631 15.42619 15.54139 15.54149
#> 10 12.25983 12.25984 12.23258 12.23268 12.23276 12.23283 12.35671 12.35665
#> 11 12.00526 12.00527 11.98186 11.98200 11.98213 11.98223 12.06533 12.06522
#> 12 12.12084 12.12085 12.09569 12.09582 12.09592 12.09601 12.23702 12.23687
#> 13 28.71109 28.71111 29.08621 29.08671 29.08711 29.08743 28.60380 28.60386
#> 14 28.46786 28.46787 28.85326 28.85347 28.85363 28.85375 30.03017 30.03033
#> 15 24.82989 24.82988 24.83824 24.83797 24.83776 24.83759 24.18978 24.18959
#> 16 17.11339 17.11337 16.99961 16.99949 16.99938 16.99929 17.26299 17.26315
#> 17 16.51355 16.51354 16.40883 16.40882 16.40881 16.40879 16.24006 16.24012
#> 18 27.69978 27.69979 26.88505 26.88499 26.88495 26.88493 26.20356 26.20307
#> 19 18.79640 18.79640 19.09582 19.09618 19.09647 19.09670 18.91671 18.91684
#> 20 19.78256 19.78256 19.59708 19.59686 19.59668 19.59654 19.43796 19.43794
#> 21 13.81482 13.81481 14.21592 14.21585 14.21579 14.21575 13.60101 13.60114
#>      0.0952   0.0888   0.0828   0.0772   0.0720   0.0671   0.0626   0.0584
#> 1  22.41880 22.41883 22.41886 22.41888 22.41890 22.41892 22.41893 22.41895
#> 2  22.08729 22.08730 22.08731 22.08732 22.08733 22.08734 22.08734 22.08736
#> 3  16.86010 16.86017 16.86023 16.86028 16.86032 16.86036 16.86039 16.86040
#> 4  20.74429 20.74448 20.74464 20.74479 20.74491 20.74502 20.74512 20.74536
#> 5  23.09850 23.09839 23.09829 23.09820 23.09813 23.09806 23.09800 23.09791
#> 6  22.93380 22.93363 22.93348 22.93335 22.93323 22.93314 22.93305 22.93291
#> 7  15.16456 15.16461 15.16465 15.16468 15.16471 15.16473 15.16475 15.16486
#> 8  15.60658 15.60665 15.60671 15.60676 15.60680 15.60684 15.60686 15.60699
#> 9  15.54157 15.54164 15.54170 15.54175 15.54179 15.54182 15.54185 15.54197
#> 10 12.35659 12.35655 12.35651 12.35647 12.35644 12.35642 12.35640 12.35630
#> 11 12.06512 12.06504 12.06498 12.06492 12.06487 12.06483 12.06480 12.06469
#> 12 12.23674 12.23663 12.23654 12.23646 12.23639 12.23633 12.23629 12.23615
#> 13 28.60392 28.60398 28.60404 28.60409 28.60414 28.60419 28.60423 28.60429
#> 14 30.03047 30.03058 30.03067 30.03074 30.03080 30.03085 30.03088 30.03078
#> 15 24.18942 24.18928 24.18916 24.18906 24.18898 24.18890 24.18884 24.18879
#> 16 17.26328 17.26339 17.26349 17.26357 17.26363 17.26368 17.26373 17.26380
#> 17 16.24017 16.24021 16.24025 16.24028 16.24031 16.24034 16.24036 16.24032
#> 18 26.20265 26.20230 26.20200 26.20175 26.20154 26.20136 26.20122 26.20093
#> 19 18.91695 18.91705 18.91714 18.91722 18.91728 18.91734 18.91739 18.91735
#> 20 19.43793 19.43792 19.43791 19.43790 19.43789 19.43789 19.43788 19.43800
#> 21 13.60126 13.60136 13.60145 13.60153 13.60159 13.60165 13.60170 13.60190
#>      0.0545   0.0508   0.0474   0.0442   0.0412   0.0384   0.0358   0.0334
#> 1  22.32031 22.32046 22.32059 22.32072 22.32084 22.32096 22.32107 22.28078
#> 2  22.19154 22.19159 22.19163 22.19166 22.19168 22.19170 22.19171 22.09533
#> 3  16.44519 16.44541 16.44563 16.44585 16.44607 16.44629 16.44651 16.39984
#> 4  20.69012 20.69021 20.69031 20.69040 20.69049 20.69059 20.69068 20.70062
#> 5  22.95233 22.95238 22.95242 22.95247 22.95251 22.95255 22.95260 23.20346
#> 6  23.07563 23.07546 23.07529 23.07513 23.07498 23.07482 23.07467 22.69572
#> 7  15.76532 15.76518 15.76502 15.76486 15.76469 15.76451 15.76432 15.88151
#> 8  15.93701 15.93700 15.93697 15.93694 15.93690 15.93686 15.93681 15.92689
#> 9  15.91176 15.91173 15.91169 15.91164 15.91158 15.91151 15.91144 15.82864
#> 10 11.95596 11.95607 11.95619 11.95632 11.95645 11.95658 11.95671 11.87655
#> 11 11.99241 11.99236 11.99231 11.99226 11.99222 11.99218 11.99214 12.00472
#> 12 12.45018 12.45001 12.44985 12.44968 12.44951 12.44935 12.44918 12.56304
#> 13 28.87613 28.87596 28.87579 28.87562 28.87545 28.87529 28.87513 28.82375
#> 14 29.85778 29.85829 29.85876 29.85919 29.85959 29.85995 29.86029 29.91514
#> 15 24.14959 24.14933 24.14910 24.14888 24.14869 24.14851 24.14835 24.26093
#> 16 17.21500 17.21521 17.21541 17.21559 17.21577 17.21593 17.21609 17.21620
#> 17 15.75868 15.75887 15.75907 15.75927 15.75948 15.75969 15.75990 15.74027
#> 18 26.21656 26.21596 26.21539 26.21487 26.21438 26.21393 26.21351 26.24849
#> 19 18.93822 18.93830 18.93838 18.93844 18.93850 18.93856 18.93860 18.98638
#> 20 19.35459 19.35454 19.35450 19.35447 19.35445 19.35444 19.35444 19.44022
#> 21 13.54568 13.54569 13.54571 13.54573 13.54577 13.54580 13.54585 13.51151
#>      0.0312   0.0291   0.0271   0.0253   0.0236   0.0220   0.0205   0.0191
#> 1  22.28087 22.28096 22.28105 22.28113 22.28121 22.33200 22.33195 22.33191
#> 2  22.09538 22.09544 22.09549 22.09554 22.09558 22.12952 22.12955 22.12958
#> 3  16.39985 16.39987 16.39989 16.39992 16.39995 16.42555 16.42531 16.42508
#> 4  20.70060 20.70059 20.70058 20.70057 20.70057 20.65172 20.65150 20.65128
#> 5  23.20362 23.20377 23.20391 23.20404 23.20416 23.06732 23.06715 23.06700
#> 6  22.69553 22.69535 22.69519 22.69503 22.69489 22.62326 22.62321 22.62317
#> 7  15.88163 15.88173 15.88182 15.88189 15.88195 15.86064 15.86087 15.86111
#> 8  15.92697 15.92704 15.92710 15.92716 15.92721 15.93301 15.93309 15.93317
#> 9  15.82869 15.82873 15.82876 15.82879 15.82882 15.83273 15.83281 15.83289
#> 10 11.87651 11.87648 11.87646 11.87644 11.87643 11.90600 11.90586 11.90571
#> 11 12.00468 12.00465 12.00461 12.00458 12.00455 12.03049 12.03057 12.03064
#> 12 12.56305 12.56305 12.56305 12.56304 12.56303 12.60873 12.60903 12.60932
#> 13 28.82362 28.82350 28.82337 28.82325 28.82313 28.84385 28.84400 28.84416
#> 14 29.91560 29.91603 29.91644 29.91682 29.91719 29.98605 29.98592 29.98579
#> 15 24.26074 24.26057 24.26040 24.26024 24.26009 24.42601 24.42640 24.42677
#> 16 17.21635 17.21650 17.21664 17.21678 17.21691 17.18520 17.18500 17.18481
#> 17 15.74026 15.74025 15.74025 15.74026 15.74027 15.75018 15.74993 15.74968
#> 18 26.24804 26.24761 26.24720 26.24680 26.24642 26.19976 26.20007 26.20036
#> 19 18.98643 18.98648 18.98652 18.98656 18.98660 18.87248 18.87233 18.87219
#> 20 19.44016 19.44011 19.44007 19.44002 19.43998 19.38413 19.38405 19.38397
#> 21 13.51140 13.51130 13.51120 13.51112 13.51104 13.55138 13.55139 13.55139
#>      0.0178   0.0166   0.0155   0.0145   0.0135   0.0126   0.0117   0.0109
#> 1  22.33186 22.33181 22.33176 22.33172 22.33167 22.33162 22.33157 22.33152
#> 2  22.12960 22.12963 22.12965 22.12968 22.12970 22.12972 22.12974 22.12975
#> 3  16.42485 16.42462 16.42439 16.42417 16.42394 16.42372 16.42351 16.42329
#> 4  20.65107 20.65087 20.65068 20.65049 20.65031 20.65014 20.64997 20.64981
#> 5  23.06687 23.06675 23.06665 23.06656 23.06648 23.06641 23.06636 23.06632
#> 6  22.62312 22.62308 22.62303 22.62298 22.62294 22.62289 22.62284 22.62280
#> 7  15.86135 15.86159 15.86183 15.86207 15.86231 15.86255 15.86279 15.86303
#> 8  15.93325 15.93333 15.93342 15.93350 15.93358 15.93367 15.93375 15.93383
#> 9  15.83297 15.83305 15.83312 15.83320 15.83328 15.83336 15.83343 15.83351
#> 10 11.90557 11.90542 11.90528 11.90513 11.90498 11.90484 11.90469 11.90454
#> 11 12.03071 12.03077 12.03084 12.03089 12.03095 12.03100 12.03105 12.03109
#> 12 12.60961 12.60989 12.61016 12.61042 12.61068 12.61093 12.61117 12.61141
#> 13 28.84430 28.84444 28.84458 28.84471 28.84483 28.84495 28.84506 28.84517
#> 14 29.98568 29.98557 29.98547 29.98538 29.98530 29.98522 29.98515 29.98509
#> 15 24.42712 24.42745 24.42775 24.42804 24.42830 24.42855 24.42878 24.42899
#> 16 17.18463 17.18447 17.18431 17.18415 17.18401 17.18388 17.18375 17.18363
#> 17 15.74944 15.74919 15.74895 15.74872 15.74848 15.74825 15.74803 15.74780
#> 18 26.20064 26.20091 26.20116 26.20140 26.20163 26.20184 26.20205 26.20224
#> 19 18.87207 18.87195 18.87185 18.87175 18.87167 18.87160 18.87153 18.87147
#> 20 19.38390 19.38383 19.38376 19.38369 19.38363 19.38357 19.38351 19.38346
#> 21 13.55139 13.55138 13.55137 13.55135 13.55133 13.55130 13.55126 13.55123
#>      0.0102   0.0095   0.0089   0.0083   0.0077   0.0072   0.0067   0.0063
#> 1  22.33147 22.33143 22.33138 22.33133 22.33129 22.33124 22.33119 22.33115
#> 2  22.12977 22.12979 22.12980 22.12982 22.12983 22.12985 22.12986 22.12987
#> 3  16.42308 16.42288 16.42267 16.42247 16.42228 16.42209 16.42190 16.42171
#> 4  20.64965 20.64951 20.64936 20.64923 20.64910 20.64897 20.64885 20.64874
#> 5  23.06629 23.06626 23.06625 23.06625 23.06626 23.06627 23.06629 23.06632
#> 6  22.62275 22.62270 22.62266 22.62261 22.62257 22.62252 22.62248 22.62243
#> 7  15.86327 15.86350 15.86374 15.86397 15.86420 15.86443 15.86466 15.86488
#> 8  15.93392 15.93400 15.93408 15.93416 15.93424 15.93433 15.93440 15.93448
#> 9  15.83358 15.83366 15.83373 15.83380 15.83388 15.83395 15.83402 15.83409
#> 10 11.90440 11.90425 11.90411 11.90396 11.90382 11.90368 11.90354 11.90341
#> 11 12.03113 12.03117 12.03121 12.03124 12.03127 12.03130 12.03133 12.03135
#> 12 12.61164 12.61186 12.61208 12.61228 12.61249 12.61268 12.61287 12.61305
#> 13 28.84528 28.84537 28.84547 28.84556 28.84564 28.84572 28.84580 28.84587
#> 14 29.98504 29.98499 29.98495 29.98491 29.98488 29.98486 29.98483 29.98482
#> 15 24.42919 24.42937 24.42954 24.42969 24.42982 24.42995 24.43006 24.43016
#> 16 17.18352 17.18341 17.18331 17.18322 17.18314 17.18306 17.18298 17.18291
#> 17 15.74758 15.74737 15.74716 15.74695 15.74675 15.74655 15.74635 15.74616
#> 18 26.20242 26.20259 26.20276 26.20291 26.20305 26.20318 26.20331 26.20342
#> 19 18.87142 18.87138 18.87135 18.87132 18.87130 18.87128 18.87127 18.87127
#> 20 19.38341 19.38336 19.38331 19.38326 19.38322 19.38318 19.38314 19.38311
#> 21 13.55119 13.55114 13.55109 13.55104 13.55099 13.55093 13.55087 13.55081
#>      0.0058   0.0054   0.0051   0.0047
#> 1  22.33110 22.33106 22.33102 22.33097
#> 2  22.12988 22.12989 22.12990 22.12991
#> 3  16.42153 16.42136 16.42118 16.42101
#> 4  20.64863 20.64853 20.64843 20.64833
#> 5  23.06635 23.06639 23.06644 23.06649
#> 6  22.62239 22.62235 22.62230 22.62226
#> 7  15.86510 15.86532 15.86553 15.86574
#> 8  15.93456 15.93464 15.93471 15.93479
#> 9  15.83415 15.83422 15.83428 15.83435
#> 10 11.90327 11.90314 11.90300 11.90287
#> 11 12.03137 12.03139 12.03141 12.03142
#> 12 12.61323 12.61340 12.61356 12.61372
#> 13 28.84594 28.84600 28.84606 28.84612
#> 14 29.98481 29.98480 29.98479 29.98479
#> 15 24.43024 24.43032 24.43039 24.43044
#> 16 17.18285 17.18279 17.18274 17.18269
#> 17 15.74598 15.74580 15.74562 15.74545
#> 18 26.20353 26.20363 26.20372 26.20380
#> 19 18.87127 18.87127 18.87128 18.87130
#> 20 19.38307 19.38304 19.38301 19.38298
#> 21 13.55075 13.55068 13.55061 13.55055
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 18.7 18.1 24.4 22.8 16.4 17.3 15.2 10.4 10.4 14.7 32.4 30.4 21.5
#> [16] 15.2 19.2 26.0 15.8 19.7 15.0
#> 
#> $X
#>               am        carb        cyl       disp         drat       gear
#>  [1,]  1.2747549  0.56842042 -0.2744106 -0.7372072  0.639377162  0.4291975
#>  [2,]  1.2747549  0.56842042 -0.2744106 -0.7372072  0.639377162  0.4291975
#>  [3,] -0.7844645 -0.62526247  0.8781141  0.8689920 -0.710514300 -0.8583951
#>  [4,] -0.7844645 -1.22210391 -0.2744106 -0.2151925 -1.412457860 -0.8583951
#>  [5,] -0.7844645 -0.62526247 -1.4269354 -0.8440195  0.261407553  0.4291975
#>  [6,] -0.7844645 -0.62526247 -1.4269354 -0.8914023  0.675374268  0.4291975
#>  [7,] -0.7844645 -0.02842102  0.8781141  0.1927822 -0.854502722 -0.8583951
#>  [8,] -0.7844645 -0.02842102  0.8781141  0.1927822 -0.854502722 -0.8583951
#>  [9,] -0.7844645 -0.02842102  0.8781141  0.1927822 -0.854502722 -0.8583951
#> [10,] -0.7844645  0.56842042  0.8781141  1.7684636 -1.106482462 -0.8583951
#> [11,] -0.7844645  0.56842042  0.8781141  1.6720916 -0.980492592 -0.8583951
#> [12,] -0.7844645  0.56842042  0.8781141  1.5114717 -0.566525877 -0.8583951
#> [13,]  1.2747549 -1.22210391 -1.4269354 -1.3901272  0.963351113  0.4291975
#> [14,]  1.2747549 -0.62526247 -1.4269354 -1.4142202  2.493228104  0.4291975
#> [15,] -0.7844645 -1.22210391 -1.4269354 -1.0576439  0.279406106 -0.8583951
#> [16,] -0.7844645 -0.62526247  0.8781141  0.4192562 -0.710514300 -0.8583951
#> [17,] -0.7844645 -0.62526247  0.8781141  1.1902319 -0.836504170 -0.8583951
#> [18,]  1.2747549 -0.62526247 -1.4269354 -1.0560377  1.593300462  1.7167902
#> [19,]  1.2747549  0.56842042  0.8781141  0.7967131  1.215330853  1.7167902
#> [20,]  1.2747549  1.76210331 -0.2744106 -0.8576721  0.135417683  1.7167902
#> [21,]  1.2747549  2.95578620  0.8781141  0.3951633 -0.008570739  1.7167902
#>                hp        qsec         vs          wt
#>  [1,] -0.63774172 -0.69314009 -0.6324555 -0.79129445
#>  [2,] -0.63774172 -0.40024295 -0.6324555 -0.53485644
#>  [3,]  0.28637447 -0.40024295 -0.6324555  0.03332976
#>  [4,] -0.70882758  1.27345500  1.5811388  0.05344254
#>  [5,] -1.32016598  1.15838826  1.5811388 -0.21808006
#>  [6,] -0.85099930  2.67517703  1.5811388 -0.25830563
#>  [7,]  0.35746033 -0.20149131 -0.6324555  0.66688251
#>  [8,]  0.35746033 -0.09688519 -0.6324555  0.32496515
#>  [9,]  0.35746033  0.11232705 -0.6324555  0.37524712
#> [10,]  0.71288963  0.10186644 -0.6324555  1.85353686
#> [11,]  0.85506135  0.01818154 -0.6324555  2.02851810
#> [12,]  1.06831893 -0.19103070 -0.6324555  1.94907260
#> [13,] -1.26329729  0.88118204  1.5811388 -1.21366295
#> [14,] -1.46233770  0.38430297  1.5811388 -1.80196194
#> [15,] -0.82256495  1.16361857  1.5811388 -0.94716854
#> [16,] -0.06905484 -0.25379438 -0.6324555  0.02830156
#> [17,]  0.28637447 -0.38455203 -0.6324555  0.44061367
#> [18,] -0.90786799 -0.56761274 -0.6324555 -1.27400131
#> [19,]  1.55170278 -1.71828008 -0.6324555 -0.23819285
#> [20,]  0.28637447 -1.19524947 -0.6324555 -0.64044856
#> [21,]  2.56112200 -1.66597702 -0.6324555  0.16406287
#> attr(,"center")
#>  [1]   0.3809524   3.0476190   6.4761905 251.7952381   3.5447619   3.6666667
#>  [7] 154.8571429  17.7852381   0.2857143   3.4068571
#> attr(,"scale")
#>  [1]   0.4856209   1.6754869   1.7353207 124.5175546   0.5556002   0.7766432
#>  [7]  70.3374761   1.9119340   0.4517540   0.9943923
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 34
#> 
#> $lambda.min
#> [1] 0.4737255
#> 
#> $null.dev
#> [1] 30.80503
#> 
#> $Bias
#> [1] 4.367272
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#>  7.68512