Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 40.11345 37.97529 35.27917 31.81503 28.95985 26.62368 24.72901 23.20908
#>   [9] 22.00646 21.07178 20.36263 19.84264 19.48066 19.25001 19.12792 19.05407
#>  [17] 18.68152 18.33243 18.24540 18.29368 18.32246 18.37140 18.46898 18.41915
#>  [25] 18.29714 18.20122 18.13966 18.66853 18.56304 18.27307 17.41833 16.99892
#>  [33] 16.69583 16.57954 16.57633 16.58104 16.58112 17.16610 18.34214 17.88898
#>  [41] 18.37590 23.18265 22.30644 29.03700 28.59679 28.54030 27.14930 27.06524
#>  [49] 27.01693 26.98609 26.12628 26.12663 26.12699 26.12152 26.83513 23.93524
#>  [57] 22.00846 22.11836 22.11784 22.08803 21.05950 21.05901 21.05853 21.05806
#>  [65] 21.05761 20.97715 21.10873 21.41911 21.41956 21.41999 21.41992 21.77533
#>  [73] 21.79439 21.81463 21.82057 21.81492 21.83621 21.83633 21.83642 21.83647
#>  [81] 21.83649 21.83646 21.83866 21.83864 21.83860 21.83854 21.83846 21.83794
#>  [89] 21.74344 21.74420 21.74495 21.74429 21.74579 21.74726 21.74868 21.75008
#>  [97] 21.75146 21.76692 21.76824 21.76953
#> 
#> $cvse
#>   [1] 11.466106 11.392132 10.951392  9.972726  9.099427  8.316315  7.612995
#>   [8]  6.982944  6.422711  5.931194  5.508922  5.157270  4.877569  4.670161
#>  [15]  4.533533  4.457619  4.349650  4.258134  4.221269  4.212940  4.199500
#>  [22]  4.190251  4.204466  4.209813  4.191794  4.182232  4.183797  4.336483
#>  [29]  4.317986  4.310540  4.335065  4.241376  4.166438  4.132400  4.123779
#>  [36]  4.122052  4.122045  4.314795  4.706515  4.768355  4.509690  5.939276
#>  [43]  5.715361  8.456726  8.453276  8.393118  7.669861  7.659687  7.657718
#>  [50]  8.300324  8.129731  8.130138  8.130542  8.176269  8.240857  6.124551
#>  [57]  6.118820  5.956065  5.955883  5.961112  5.890004  5.889890  5.889791
#>  [64]  5.889703  5.889628  6.008365  6.012772  6.076786  6.077099  6.077394
#>  [71]  6.077534  6.182498  6.167485  6.164105  6.163221  6.164315  6.162051
#>  [78]  6.162125  6.162185  6.162234  6.162270  6.162299  6.161932  6.161930
#>  [85]  6.161919  6.161898  6.161869  6.161694  6.133987  6.134232  6.134472
#>  [92]  6.142898  6.143472  6.144027  6.144564  6.145080  6.145577  6.143850
#>  [99]  6.144309  6.144751
#> 
#> $fold
#>  [1]  8  1  7  5  5  6  4  2  4  8  1  9  2  3 10 10  6  3  9  1  7
#> 
#> $lambda
#>   [1] 5.315698728 4.957438424 4.623323667 4.311727125 4.021131148 3.750120367
#>   [7] 3.497374805 3.261663449 3.041838249 2.836828531 2.645635783 2.467328786
#>  [13] 2.301039083 2.145956751 2.001326448 1.866443743 1.740651682 1.623337584
#>  [19] 1.513930064 1.411896245 1.316739163 1.227995351 1.145232574 1.068047732
#>  [25] 0.996064889 0.928933449 0.866326444 0.807938941 0.753486560 0.702704088
#>  [31] 0.655344184 0.611176180 0.569984951 0.531569873 0.495743842 0.462332367
#>  [37] 0.431172713 0.402113115 0.375012037 0.349737481 0.326166345 0.304183825
#>  [43] 0.283682853 0.264563578 0.246732879 0.230103908 0.214595675 0.200132645
#>  [49] 0.186644374 0.174065168 0.162333758 0.151393006 0.141189624 0.131673916
#>  [55] 0.122799535 0.114523257 0.106804773 0.099606489 0.092893345 0.086632644
#>  [61] 0.080793894 0.075348656 0.070270409 0.065534418 0.061117618 0.056998495
#>  [67] 0.053156987 0.049574384 0.046233237 0.043117271 0.040211311 0.037501204
#>  [73] 0.034973748 0.032616634 0.030418382 0.028368285 0.026456358 0.024673288
#>  [79] 0.023010391 0.021459568 0.020013264 0.018664437 0.017406517 0.016233376
#>  [85] 0.015139301 0.014118962 0.013167392 0.012279954 0.011452326 0.010680477
#>  [91] 0.009960649 0.009289334 0.008663264 0.008079389 0.007534866 0.007027041
#>  [97] 0.006553442 0.006111762 0.005699850 0.005315699
#> 
#> $fit
#> $beta
#>               5.3157     4.9574    4.6233   4.3117    4.0211    3.7501
#> (Intercept) 21.15714 22.8034521 24.338806 25.77068 27.106053 28.351426
#> am           0.00000  0.0000000  0.000000  0.00000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.000000  0.00000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.000000  0.00000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.000000  0.00000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.000000  0.00000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.000000  0.00000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.000000  0.00000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.000000  0.00000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.000000  0.00000  0.000000  0.000000
#> wt           0.00000 -0.5249232 -1.014468 -1.47102 -1.896801 -2.293886
#>                3.4974    3.2617    3.0418    2.8368    2.6456    2.4673
#> (Intercept) 29.512865 30.596026 31.606186 32.548265 33.426851 34.246223
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.664209 -3.009574 -3.331662 -3.632042 -3.912178 -4.173433
#>                2.3010    2.1460   2.0013    1.8664    1.7407    1.6233
#> (Intercept) 35.010372 35.723020 36.38764 37.007462 37.441920 37.441920
#> am           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> wt          -4.417081 -4.644308 -4.85622 -5.053851 -5.192377 -5.192377
#>                1.5139    1.4119    1.3167      1.2280     1.1452     1.0680
#> (Intercept) 37.441920 37.441920 37.441920 36.10908546 34.2481736 32.5126357
#> am           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> carb         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> cyl          0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> disp         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> qsec         0.000000  0.000000  0.000000  0.06963329  0.1668557  0.2575281
#> vs           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> wt          -5.192377 -5.192377 -5.192377 -5.16860313 -5.1354102 -5.1044536
#>                 0.9961     0.9289     0.8663    0.8079     0.7535     0.7027
#> (Intercept) 30.8940672 29.3845847 27.9768364 26.663966 25.4395779 11.0656006
#> am           0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.000000  0.0000000  1.6352603
#> hp           0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> qsec         0.3420894  0.4209516  0.4944987  0.563089  0.6270565  0.9437942
#> vs           0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> wt          -5.0755834 -5.0486589 -5.0235490 -5.000132 -4.9782922 -4.1070569
#>                0.6553    0.6112    0.5700    0.5316    0.4957    0.4623
#> (Intercept) -1.385176 -2.849647 -3.048861 -3.052654 -3.054964 -3.056370
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         3.042039  3.100241  3.108046  3.108505  3.108784  3.108955
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         1.220726  1.281781  1.290117  1.290193  1.290240  1.290268
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -3.356033 -3.308028 -3.301540 -3.301300 -3.301154 -3.301065
#>                0.4312    0.4021    0.3750    0.3497    0.3262    0.3042
#> (Intercept) -3.057226 -3.057747 -3.058064 -3.058257 -3.058375 -3.058447
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         3.109058  3.109121  3.109160  3.109183  3.109197  3.109206
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         1.290285  1.290296  1.290302  1.290306  1.290308  1.290310
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -3.301010 -3.300977 -3.300957 -3.300945 -3.300937 -3.300933
#>                 0.2837     0.2646     0.2467    0.2301     0.2146     0.2001
#> (Intercept)  4.1183585  4.1175610  4.1169402  3.392128  3.3916940 -54.627941
#> am           0.0000000  0.0000000  0.0000000  1.615097  1.6144163   0.000000
#> carb        -1.0011628 -1.0014516 -1.0016761 -0.971583 -0.9718646  -1.946031
#> cyl          0.0000000  0.0000000  0.0000000  0.000000  0.0000000   2.847762
#> disp         0.0000000  0.0000000  0.0000000  0.000000  0.0000000   0.000000
#> drat         0.0000000  0.0000000  0.0000000  0.000000  0.0000000   4.639657
#> gear         3.6131963  3.6136876  3.6140695  2.994107  2.9948253   5.645079
#> hp           0.0000000  0.0000000  0.0000000  0.000000  0.0000000   0.000000
#> qsec         0.7861641  0.7860767  0.7860087  0.855903  0.8557921   1.996263
#> vs           0.0000000  0.0000000  0.0000000  0.000000  0.0000000   0.000000
#> wt          -2.4908877 -2.4904729 -2.4901503 -2.190360 -2.1901004  -2.965256
#>                 0.1866       0.1741       0.1623       0.1514       0.1412
#> (Intercept) -54.633229 -85.63354904 -85.63242588 -85.63144387 -85.63059410
#> am            0.000000   1.43006905   1.42989339   1.42972322   1.42955845
#> carb         -1.946124  -0.68637192  -0.68659906  -0.68681904  -0.68703198
#> cyl           2.848065   2.62851731   2.62866123   2.62880419   2.62894599
#> disp          0.000000   0.03838882   0.03838407   0.03837952   0.03837515
#> drat          4.639878   5.19425861   5.19439615   5.19453379   5.19467125
#> gear          5.645377   5.97005825   5.97016987   5.97028282   5.97039675
#> hp            0.000000   0.00000000   0.00000000   0.00000000   0.00000000
#> qsec          1.996386   3.55300468   3.55285557   3.55271438   3.55258079
#> vs            0.000000   0.00000000   0.00000000   0.00000000   0.00000000
#> wt           -2.965366  -6.43384928  -6.43337982  -6.43293025  -6.43249995
#>                   0.1317       0.1228       0.1145       0.1068       0.0996
#> (Intercept) -85.62986811 -85.62925781 -85.62875554 -85.62835397 -85.62804616
#> am            1.42939902   1.42924482   1.42909576   1.42895175   1.42881269
#> carb         -0.68723799  -0.68743718  -0.68762969  -0.68781564  -0.68799515
#> cyl           2.62908643   2.62922533   2.62936253   2.62949788   2.62963126
#> disp          0.03837096   0.03836695   0.03836311   0.03835944   0.03835592
#> drat          5.19480827   5.19494459   5.19508001   5.19521432   5.19534734
#> gear          5.97051137   5.97062639   5.97074156   5.97085662   5.97097135
#> hp            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> qsec          3.55245447   3.55233512   3.55222244   3.55211615   3.55201597
#> vs            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> wt           -6.43208831  -6.43169471  -6.43131855  -6.43095925  -6.43061623
#>                   0.0929       0.0866       0.0808       0.0753       0.0703
#> (Intercept) -85.62782551 -85.62768575 -85.62762093 -85.62762541 -85.62769385
#> am            1.42867848   1.42854901   1.42842417   1.42830388   1.42818800
#> carb         -0.68816837  -0.68833543  -0.68849647  -0.68865163  -0.68880105
#> cyl           2.62976255   2.62989164   2.63001845   2.63014289   2.63026490
#> disp          0.03835256   0.03834935   0.03834629   0.03834336   0.03834057
#> drat          5.19547889   5.19560882   5.19573701   5.19586331   5.19598764
#> gear          5.97108555   5.97119903   5.97131162   5.97142316   5.97153350
#> hp            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> qsec          3.55192162   3.55183286   3.55174943   3.55167108   3.55159759
#> vs            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> wt           -6.43028894  -6.42997681  -6.42967931  -6.42939591  -6.42912609
#>                   0.0655       0.0611       0.0570       0.0532       0.0496
#> (Intercept) -85.62782117 -85.62800259 -85.62823358 -85.62850985 -85.62882736
#> am            1.42807645   1.42796910   1.42786586   1.42776661   1.42767125
#> carb         -0.68894486  -0.68908322  -0.68921626  -0.68934413  -0.68946697
#> cyl           2.63038442   2.63050139   2.63061579   2.63072758   2.63083673
#> disp          0.03833791   0.03833537   0.03833295   0.03833065   0.03832847
#> drat          5.19610987   5.19622994   5.19634777   5.19646329   5.19657645
#> gear          5.97164251   5.97175008   5.97185611   5.97196049   5.97206315
#> hp            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> qsec          3.55152872   3.55146426   3.55140400   3.55134773   3.55129527
#> vs            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> wt           -6.42886935  -6.42862521  -6.42839317  -6.42817279  -6.42796359
#>                   0.0462       0.0431       0.0402       0.0375       0.0350
#> (Intercept) -85.62918230 -85.62957106 -85.62999027 -85.63043672 -85.63090743
#> am            1.42757967   1.42749176   1.42740741   1.42732652   1.42724900
#> carb         -0.68958491  -0.68969810  -0.68980667  -0.68991077  -0.69001052
#> cyl           2.63094323   2.63104706   2.63114823   2.63124674   2.63134259
#> disp          0.03832639   0.03832441   0.03832253   0.03832075   0.03831906
#> drat          5.19668720   5.19679551   5.19690134   5.19700467   5.19710549
#> gear          5.97216401   5.97226300   5.97236007   5.97245518   5.97254828
#> hp            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> qsec          3.55124642   3.55120101   3.55115886   3.55111980   3.55108368
#> vs            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> wt           -6.42776515  -6.42757703  -6.42739881  -6.42723010  -6.42707050
#>                   0.0326       0.0304       0.0284       0.0265        0.0247
#> (Intercept) -85.63139958 -85.63191053 -85.63243779 -85.63297906 -89.463584280
#> am            1.42717473   1.42710361   1.42703554   1.42697043   1.496451525
#> carb         -0.69010607  -0.69019754  -0.69028506  -0.69036876  -0.701695384
#> cyl           2.63143579   2.63152637   2.63161433   2.63169971   2.661630632
#> disp          0.03831746   0.03831595   0.03831451   0.03831315   0.037314365
#> drat          5.19720378   5.19729955   5.19739280   5.19748353   5.182088375
#> gear          5.97263934   5.97272834   5.97281525   5.97290006   6.160120136
#> hp            0.00000000   0.00000000   0.00000000   0.00000000   0.006249304
#> qsec          3.55105034   3.55101964   3.55099143   3.55096558   3.693862917
#> vs            0.00000000   0.00000000   0.00000000   0.00000000   0.000000000
#> wt           -6.42691963  -6.42677712  -6.42664261  -6.42651576  -6.479825767
#>                    0.0230       0.0215        0.0200        0.0187
#> (Intercept) -89.466255721 -89.46885405 -90.563867485 -90.572262843
#> am            1.496215650   1.49598913   1.449135584   1.449108233
#> carb         -0.701985386  -0.70226161  -0.691409766  -0.691396013
#> cyl           2.661970503   2.66229832   2.622899802   2.622939335
#> disp          0.037310335   0.03730657   0.037253463   0.037252451
#> drat          5.182451377   5.18280429   5.140013565   5.139956960
#> gear          6.160460998   6.16078703   6.208708198   6.209157260
#> hp            0.006248344   0.00624730   0.007656654   0.007666358
#> qsec          3.693792584   3.69372661   3.774957544   3.775320179
#> vs            0.000000000   0.00000000  -0.266010307  -0.266445621
#> wt           -6.479438748  -6.47907538  -6.544638210  -6.544853478
#>                    0.0174        0.0162        0.0151       0.0141
#> (Intercept) -90.580277180 -90.587927592 -90.595230422 -90.60220128
#> am            1.449081090   1.449054231   1.449027724   1.44900163
#> carb         -0.691385243  -0.691377191  -0.691371607  -0.69136826
#> cyl           2.622980822   2.623024020   2.623068702   2.62311465
#> disp          0.037251468   0.037250516   0.037249595   0.03724871
#> drat          5.139907346   5.139864261   5.139827264   5.13979593
#> gear          6.209587391   6.209999332   6.210393800   6.21077149
#> hp            0.007675563   0.007684291   0.007692566   0.00770041
#> qsec          3.775664224   3.775990606   3.776300208   3.77659387
#> vs           -0.266856168  -0.267243209  -0.267607949  -0.26795154
#> wt           -6.545055656  -6.545245563  -6.545423969  -6.54559160
#>                    0.0132        0.0123        0.0115        0.0107
#> (Intercept) -90.608855098 -90.615206122 -90.621267977 -90.627053676
#> am            1.448975992   1.448950861   1.448926274   1.448902261
#> carb         -0.691366937  -0.691367433  -0.691369560  -0.691373143
#> cyl           2.623161674   2.623209575   2.623258181   2.623307328
#> disp          0.037247846   0.037247019   0.037246224   0.037245459
#> drat          5.139769865   5.139748677   5.139732006   5.139719506
#> gear          6.211133057   6.211479157   6.211810406   6.212127399
#> hp            0.007707843   0.007714885   0.007721555   0.007727871
#> qsec          3.776872398   3.777136553   3.777387062   3.777624617
#> vs           -0.268275078  -0.268579618  -0.268866160  -0.269135660
#> wt           -6.545749136  -6.545897224  -6.546036468  -6.546167437
#>                    0.0100        0.0093        0.0087        0.0081
#> (Intercept) -90.632575652 -90.637845783 -90.642875418 -90.647675402
#> am            1.448878852   1.448856068   1.448833928   1.448812448
#> carb         -0.691378018  -0.691384034  -0.691391049  -0.691398933
#> cyl           2.623356863   2.623406644   2.623456540   2.623506428
#> disp          0.037244726   0.037244024   0.037243353   0.037242712
#> drat          5.139710849   5.139705725   5.139703841   5.139704920
#> gear          6.212430711   6.212720893   6.212998477   6.213263973
#> hp            0.007733851   0.007739511   0.007744867   0.007749935
#> qsec          3.777849879   3.778063474   3.778265999   3.778458022
#> vs           -0.269389029  -0.269627137  -0.269850811  -0.270060838
#> wt           -6.546290666  -6.546406659  -6.546515887  -6.546618794
#>                    0.0075       0.0070        0.0066        0.0061
#> (Intercept) -90.652256094 -90.65662739 -90.660798759 -90.664779227
#> am            1.448791640   1.44877151   1.448752070   1.448733318
#> carb         -0.691407563  -0.69141683  -0.691426624  -0.691436856
#> cyl           2.623556195   2.62360574   2.623654963   2.623703779
#> disp          0.037242100   0.03724152   0.037240964   0.037240439
#> drat          5.139708700   5.13971494   5.139723396   5.139733862
#> gear          6.213517872   6.21376065   6.213992750   6.214214619
#> hp            0.007754728   0.00775926   0.007763545   0.007767595
#> qsec          3.778640082   3.77881269   3.778976347   3.779131504
#> vs           -0.270257970  -0.27044292  -0.270616367  -0.270778957
#> wt           -6.546715797  -6.54680729  -6.546893627  -6.546975164
#>                    0.0057        0.0053
#> (Intercept) -90.668577433 -90.672201627
#> am            1.448715257   1.448697887
#> carb         -0.691447435  -0.691458282
#> cyl           2.623752109   2.623799880
#> disp          0.037239941   0.037239470
#> drat          5.139746130   5.139760008
#> gear          6.214426672   6.214629314
#> hp            0.007771422   0.007775037
#> qsec          3.779278607   3.779418075
#> vs           -0.270931303  -0.271073989
#> wt           -6.547052218  -6.547125090
#> 
#> $iter
#>   [1]   0   4   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   5   4   4   4   4   4   4   4  63  31  10   5   1   1   1
#>  [37]   1   1   1   1   1   1  36   1   1  55   1 143   1 207   1   1   1   1
#>  [55]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1  72   1   1  57   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.315698728 4.957438424 4.623323667 4.311727125 4.021131148 3.750120367
#>   [7] 3.497374805 3.261663449 3.041838249 2.836828531 2.645635783 2.467328786
#>  [13] 2.301039083 2.145956751 2.001326448 1.866443743 1.740651682 1.623337584
#>  [19] 1.513930064 1.411896245 1.316739163 1.227995351 1.145232574 1.068047732
#>  [25] 0.996064889 0.928933449 0.866326444 0.807938941 0.753486560 0.702704088
#>  [31] 0.655344184 0.611176180 0.569984951 0.531569873 0.495743842 0.462332367
#>  [37] 0.431172713 0.402113115 0.375012037 0.349737481 0.326166345 0.304183825
#>  [43] 0.283682853 0.264563578 0.246732879 0.230103908 0.214595675 0.200132645
#>  [49] 0.186644374 0.174065168 0.162333758 0.151393006 0.141189624 0.131673916
#>  [55] 0.122799535 0.114523257 0.106804773 0.099606489 0.092893345 0.086632644
#>  [61] 0.080793894 0.075348656 0.070270409 0.065534418 0.061117618 0.056998495
#>  [67] 0.053156987 0.049574384 0.046233237 0.043117271 0.040211311 0.037501204
#>  [73] 0.034973748 0.032616634 0.030418382 0.028368285 0.026456358 0.024673288
#>  [79] 0.023010391 0.021459568 0.020013264 0.018664437 0.017406517 0.016233376
#>  [85] 0.015139301 0.014118962 0.013167392 0.012279954 0.011452326 0.010680477
#>  [91] 0.009960649 0.009289334 0.008663264 0.008079389 0.007534866 0.007027041
#>  [97] 0.006553442 0.006111762 0.005699850 0.005315699
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 29
#> 
#> $loss
#>   [1] 790.19143 676.27855 580.97348 501.59842 435.84148 381.70786 337.47748
#>   [8] 301.66820 273.00388 250.38670 232.87307 219.65285 210.03118 203.41288
#>  [15] 199.28880 197.22407 196.80172 196.80172 196.80172 196.80172 196.80172
#>  [22] 191.69345 185.19413 179.79736 175.34242 171.69054 168.72209 166.33407
#>  [29] 164.43780 133.06229 123.55325 123.38512 123.38256 123.38256 123.38255
#>  [36] 123.38255 123.38255 123.38255 123.38255 123.38255 123.38255 123.38255
#>  [43] 115.40709 115.40708 115.40707 111.51607 111.51606  85.19671  85.19670
#>  [50]  65.75321  65.75320  65.75318  65.75317  65.75315  65.75314  65.75313
#>  [57]  65.75312  65.75311  65.75310  65.75310  65.75309  65.75308  65.75307
#>  [64]  65.75307  65.75306  65.75306  65.75305  65.75305  65.75305  65.75304
#>  [71]  65.75304  65.75304  65.75303  65.75303  65.75303  65.75303  65.75303
#>  [78]  65.55136  65.55134  65.55133  65.49222  65.49221  65.49219  65.49218
#>  [85]  65.49217  65.49216  65.49215  65.49214  65.49213  65.49212  65.49212
#>  [92]  65.49211  65.49210  65.49210  65.49209  65.49209  65.49209  65.49208
#>  [99]  65.49208  65.49208
#> 
#> $linear.predictors
#>      5.3157   4.9574   4.6233   4.3117   4.0211   3.7501   3.4974   3.2617
#> 1  21.15714 21.42815 21.68090 21.91661 22.13643 22.34144 22.53264 22.71094
#> 2  21.15714 21.29430 21.42221 21.54150 21.65275 21.75650 21.85326 21.94350
#> 3  21.15714 21.58563 21.98524 22.35792 22.70547 23.02961 23.33190 23.61381
#> 4  21.15714 21.11582 21.07729 21.04135 21.00784 20.97658 20.94743 20.92025
#> 5  21.15714 20.98722 20.82874 20.68095 20.54312 20.41458 20.29470 20.18290
#> 6  21.15714 20.92948 20.71715 20.51914 20.33447 20.16225 20.00164 19.85185
#> 7  21.15714 21.12895 21.10265 21.07813 21.05526 21.03393 21.01404 20.99549
#> 8  21.15714 20.99772 20.84903 20.71037 20.58106 20.46046 20.34798 20.24309
#> 9  21.15714 20.66701 20.20992 19.78363 19.38607 19.01531 18.66953 18.34706
#> 10 21.15714 20.81924 20.50412 20.21023 19.93614 19.68054 19.44215 19.21984
#> 11 21.15714 19.95627 18.83633 17.79187 16.81780 15.90939 15.06219 14.27210
#> 12 21.15714 19.99774 18.91647 17.90808 16.96765 16.09060 15.27267 14.50985
#> 13 21.15714 21.64862 22.10698 22.53444 22.93309 23.30488 23.65160 23.97496
#> 14 21.15714 21.84022 22.47726 23.07136 23.62542 24.14214 24.62404 25.07346
#> 15 21.15714 21.50952 21.83814 22.14462 22.43044 22.69700 22.94559 23.17743
#> 16 21.15714 20.78775 20.44325 20.12197 19.82234 19.54290 19.28230 19.03926
#> 17 21.15714 20.78512 20.43817 20.11461 19.81285 19.53143 19.26898 19.02421
#> 18 21.15714 21.78773 22.37581 22.92426 23.43574 23.91276 24.35762 24.77250
#> 19 21.15714 21.68012 22.16784 22.62270 23.04690 23.44251 23.81146 24.15554
#> 20 21.15714 22.00924 22.80391 23.54503 24.23619 24.88078 25.48192 26.04254
#> 21 21.15714 21.34417 21.51858 21.68125 21.83295 21.97442 22.10636 22.22941
#>      3.0418   2.8368   2.6456   2.4673   2.3010   2.1460   2.0013    1.8664
#> 1  22.87723 23.03231 23.17694 23.31183 23.43762 23.55493 23.66434 23.766374
#> 2  22.02766 22.10614 22.17934 22.24760 22.31126 22.37063 22.42600 22.477642
#> 3  23.87673 24.12193 24.35060 24.56386 24.76274 24.94823 25.12121 25.282529
#> 4  20.89489 20.87125 20.84920 20.82863 20.80946 20.79157 20.77489 20.759333
#> 5  20.07864 19.98140 19.89071 19.80614 19.72727 19.65371 19.58511 19.521139
#> 6  19.71215 19.58187 19.46038 19.34707 19.24139 19.14284 19.05093 18.965216
#> 7  20.97819 20.96205 20.94700 20.93297 20.91988 20.90768 20.89629 20.885679
#> 8  20.14527 20.05404 19.96896 19.88961 19.81561 19.74660 19.68224 19.622216
#> 9  18.04632 17.76585 17.50429 17.26035 17.03285 16.82069 16.62282 16.438290
#> 10 19.01250 18.81915 18.63882 18.47064 18.31380 18.16754 18.03112 17.903907
#> 11 13.53525 12.84807 12.20720 11.60952 11.05212 10.53229 10.04750  9.595377
#> 12 13.79845 13.13500 12.51626 11.93922 11.40107 10.89919 10.43114  9.994631
#> 13 24.27653 24.55777 24.82006 25.06467 25.29279 25.50554 25.70395 25.888991
#> 14 25.49259 25.88347 26.24800 26.58797 26.90503 27.20071 27.47647 27.733646
#> 15 23.39364 23.59528 23.78333 23.95871 24.12227 24.27480 24.41705 24.549721
#> 16 18.81260 18.60122 18.40409 18.22024 18.04878 17.88888 17.73975 17.600676
#> 17 18.79595 18.58306 18.38453 18.19937 18.02669 17.86566 17.71547 17.575407
#> 18 25.15942 25.52026 25.85679 26.17063 26.46332 26.73628 26.99085 27.228261
#> 19 24.47643 24.77569 25.05479 25.31507 25.55782 25.78420 25.99533 26.192222
#> 20 26.56538 27.05299 27.50773 27.93182 28.32733 28.69618 29.04018 29.360986
#> 21 22.34417 22.45119 22.55100 22.64408 22.73089 22.81184 22.88734 22.957758
#>       1.7407    1.6233    1.5139    1.4119    1.3167    1.2280    1.1452
#> 1  23.837893 23.837893 23.837893 23.837893 23.837893 23.713509 23.539844
#> 2  22.513837 22.513837 22.513837 22.513837 22.513837 22.434510 22.323754
#> 3  25.395606 25.395606 25.395606 25.395606 25.395606 25.413802 25.439207
#> 4  20.748429 20.748429 20.748429 20.748429 20.748429 20.845698 20.981505
#> 5  19.476296 19.476296 19.476296 19.476296 19.476296 19.633704 19.853477
#> 6  18.905135 18.905135 18.905135 18.905135 18.905135 18.760164 18.557754
#> 7  20.878238 20.878238 20.878238 20.878238 20.878238 21.013907 21.203330
#> 8  19.580144 19.580144 19.580144 19.580144 19.580144 19.603380 19.635822
#> 9  16.308946 16.308946 16.308946 16.308946 16.308946 16.284490 16.250344
#> 10 17.814736 17.814736 17.814736 17.814736 17.814736 17.825165 17.839726
#> 11  9.278468  9.278468  9.278468  9.278468  9.278468  9.315447  9.367078
#> 12  9.688666  9.688666  9.688666  9.688666  9.688666  9.695914  9.706033
#> 13 26.018691 26.018691 26.018691 26.018691 26.018691 26.093919 26.198952
#> 14 27.913909 27.913909 27.913909 27.913909 27.913909 28.010401 28.145125
#> 15 24.642711 24.642711 24.642711 24.642711 24.642711 24.761841 24.928171
#> 16 17.503193 17.503193 17.503193 17.503193 17.503193 17.334698 17.099445
#> 17 17.477231 17.477231 17.477231 17.477231 17.477231 17.423054 17.347411
#> 18 27.394671 27.394671 27.394671 27.394671 27.394671 27.423908 27.464728
#> 19 26.330234 26.330234 26.330234 26.330234 26.330234 26.211151 26.044886
#> 20 29.585854 29.585854 29.585854 29.585854 29.585854 29.465792 29.298160
#> 21 23.007113 23.007113 23.007113 23.007113 23.007113 23.035548 23.075250
#>       1.0680    0.9961    0.9289    0.8663    0.8079    0.7535   0.7027
#> 1  23.377879 23.226830 23.085961 22.954587 22.832066 22.717803 22.38101
#> 2  22.220459 22.124126 22.034286 21.950501 21.872362 21.799490 21.86223
#> 3  25.462901 25.484997 25.505605 25.524823 25.542747 25.559462 25.64228
#> 4  21.108163 21.226284 21.336445 21.439181 21.534993 21.624348 21.11455
#> 5  20.058444 20.249596 20.427866 20.594120 20.749170 20.893770 20.84448
#> 6  18.368981 18.192930 18.028745 17.875626 17.732826 17.599650 16.25889
#> 7  21.379990 21.544744 21.698394 21.841689 21.975326 22.099957 23.38101
#> 8  19.666079 19.694296 19.720612 19.745154 19.768042 19.789387 20.74980
#> 9  16.218498 16.188798 16.161100 16.135269 16.111179 16.088712 15.67768
#> 10 17.853306 17.865971 17.877782 17.888798 17.899071 17.908651 17.43500
#> 11  9.415229  9.460136  9.502016  9.541073  9.577498  9.611468 10.51312
#> 12  9.715470  9.724271  9.732479  9.740134  9.747273  9.753931 10.46006
#> 13 26.296909 26.388264 26.473462 26.552918 26.627019 26.696126 26.94679
#> 14 28.270772 28.387951 28.497232 28.599148 28.694195 28.782837 28.85170
#> 15 25.083294 25.227963 25.362882 25.488707 25.606052 25.715489 24.73281
#> 16 16.880041 16.675425 16.484598 16.306633 16.140662 15.985877 14.74415
#> 17 17.276865 17.211073 17.149716 17.092493 17.039127 16.989358 16.27144
#> 18 27.502798 27.538303 27.571414 27.602294 27.631093 27.657951 27.49720
#> 19 25.889824 25.745212 25.610346 25.484570 25.367271 25.257877 26.21416
#> 20 29.141822 28.996020 28.860045 28.733235 28.614971 28.504677 28.97805
#> 21 23.112277 23.146808 23.179012 23.209046 23.237055 23.263177 23.74360
#>      0.6553   0.6112   0.5700   0.5316   0.4957   0.4623   0.4312   0.4021
#> 1  22.08332 21.98241 21.96861 21.96854 21.96849 21.96847 21.96845 21.96844
#> 2  21.91114 21.85666 21.84918 21.84921 21.84923 21.84925 21.84925 21.84926
#> 3  25.71469 25.73065 25.73282 25.73284 25.73286 25.73286 25.73287 25.73287
#> 4  20.68220 20.73360 20.74070 20.74054 20.74044 20.74038 20.74034 20.74032
#> 5  20.81214 20.92292 20.93811 20.93807 20.93804 20.93803 20.93802 20.93801
#> 6  15.09620 14.94483 14.92423 14.92388 14.92367 14.92354 14.92346 14.92341
#> 7  24.49175 24.63434 24.65375 24.65408 24.65428 24.65441 24.65448 24.65453
#> 8  21.57751 21.62830 21.63516 21.63543 21.63559 21.63569 21.63575 21.63578
#> 9  15.32251 15.29040 15.28604 15.28593 15.28586 15.28582 15.28580 15.28578
#> 10 17.02820 17.01880 17.01756 17.01742 17.01734 17.01729 17.01726 17.01724
#> 11 11.29115 11.34968 11.35760 11.35785 11.35800 11.35809 11.35815 11.35818
#> 12 11.06799 11.09830 11.10238 11.10258 11.10270 11.10277 11.10281 11.10284
#> 13 27.16724 27.22994 27.23851 27.23857 27.23860 27.23862 27.23863 27.23864
#> 14 28.91710 28.98854 28.99832 28.99832 28.99832 28.99832 28.99833 28.99833
#> 15 23.89504 23.94523 23.95222 23.95192 23.95174 23.95163 23.95156 23.95152
#> 16 13.66516 13.50050 13.47806 13.47774 13.47755 13.47743 13.47736 13.47732
#> 17 15.65037 15.58608 15.57735 15.57715 15.57704 15.57697 15.57692 15.57690
#> 18 27.36077 27.37595 27.37805 27.37800 27.37797 27.37795 27.37794 27.37793
#> 19 27.02923 26.97813 26.97102 26.97131 26.97149 26.97160 26.97166 26.97170
#> 20 29.37761 29.30862 29.29911 29.29927 29.29936 29.29942 29.29945 29.29947
#> 21 24.15871 24.19613 24.20121 24.20134 24.20142 24.20147 24.20150 24.20152
#>      0.3750   0.3497   0.3262   0.3042   0.2837   0.2646   0.2467   0.2301
#> 1  21.96844 21.96843 21.96843 21.96843 20.98063 20.98029 20.98002 21.44674
#> 2  21.84926 21.84926 21.84926 21.84926 20.78570 20.78542 20.78520 21.36751
#> 3  25.73287 25.73287 25.73287 25.73287 26.42164 26.42185 26.42202 26.85879
#> 4  20.74031 20.74030 20.74029 20.74029 21.23161 21.23163 21.23165 20.99961
#> 5  20.93801 20.93800 20.93800 20.93800 21.23455 21.23461 21.23465 21.13058
#> 6  14.92338 14.92336 14.92335 14.92334 14.51367 14.51328 14.51299 14.22604
#> 7  24.65456 24.65458 24.65459 24.65459 24.34617 24.34633 24.34646 23.55620
#> 8  21.63581 21.63582 21.63583 21.63583 20.38464 20.38448 20.38436 19.61041
#> 9  15.28577 15.28577 15.28576 15.28576 15.49580 15.49578 15.49576 15.43765
#> 10 17.01723 17.01722 17.01722 17.01722 16.68986 16.68966 16.68951 16.58639
#> 11 11.35820 11.35822 11.35822 11.35823 11.45217 11.45238 11.45254 11.85980
#> 12 11.10286 11.10287 11.10287 11.10288 11.33448 11.33470 11.33486 11.69047
#> 13 27.23865 27.23865 27.23865 27.23865 27.39664 27.39673 27.39680 27.85771
#> 14 28.99833 28.99833 28.99833 28.99833 28.64387 28.64377 28.64369 29.02523
#> 15 23.95150 23.95148 23.95147 23.95147 23.54789 23.54755 23.54729 23.13025
#> 16 13.47729 13.47728 13.47727 13.47726 13.50308 13.50284 13.50266 13.26660
#> 17 15.57688 15.57687 15.57687 15.57686 16.78226 16.78246 16.78262 16.60250
#> 18 27.37793 27.37793 27.37792 27.37792 27.60862 27.60864 27.60867 27.95029
#> 19 26.97173 26.97174 26.97175 26.97176 27.98046 27.98096 27.98136 27.64080
#> 20 29.29949 29.29950 29.29950 29.29950 29.69948 29.69971 29.69989 29.18534
#> 21 24.20153 24.20154 24.20154 24.20154 24.26680 24.26692 24.26701 24.87108
#>      0.2146   0.2001   0.1866   0.1741   0.1623   0.1514   0.1412   0.1317
#> 1  21.44623 20.43900 20.43895 20.72796 20.72786 20.72777 20.72767 20.72758
#> 2  21.36700 20.80077 20.80076 21.07701 21.07695 21.07689 21.07683 21.07677
#> 3  26.85881 25.53113 25.53104 24.84323 24.84330 24.84337 24.84343 24.84349
#> 4  20.99973 21.01203 21.01208 21.64958 21.64948 21.64939 21.64930 21.64922
#> 5  21.13068 20.35794 20.35799 19.91563 19.91565 19.91566 19.91568 19.91569
#> 6  14.22580 13.23341 13.23333 14.36357 14.36332 14.36308 14.36284 14.36261
#> 7  23.55669 23.03779 23.03764 22.72258 22.72260 22.72261 22.72263 22.72264
#> 8  19.61059 21.77341 21.77350 20.95530 20.95546 20.95561 20.95576 20.95590
#> 9  15.43765 16.16143 16.16155 13.41617 13.41653 13.41687 13.41721 13.41752
#> 10 16.58626 18.21911 18.21934 17.41379 17.41392 17.41405 17.41418 17.41430
#> 11 11.85982 10.71410 10.71401 12.21826 12.21807 12.21790 12.21773 12.21756
#> 12 11.69052 11.21697 11.21689 11.73223 11.73220 11.73217 11.73214 11.73210
#> 13 27.85760 28.67087 28.67095 28.74076 28.74082 28.74088 28.74093 28.74099
#> 14 29.02498 31.26113 31.26133 33.05235 33.05223 33.05211 33.05201 33.05191
#> 15 23.13011 21.55491 21.55464 21.16976 21.16968 21.16960 21.16952 21.16943
#> 16 13.26648 13.98702 13.98697 13.41577 13.41583 13.41588 13.41593 13.41598
#> 17 16.60276 18.12235 18.12254 19.12644 19.12639 19.12633 19.12629 19.12625
#> 18 27.95017 28.31879 28.31883 28.43203 28.43205 28.43207 28.43209 28.43211
#> 19 27.64142 28.64206 28.64209 27.98362 27.98380 27.98397 27.98414 27.98431
#> 20 29.18577 27.83836 27.83833 28.33263 28.33252 28.33241 28.33230 28.33219
#> 21 24.87094 23.40743 23.40725 23.01131 23.01135 23.01139 23.01142 23.01144
#>      0.1228   0.1145   0.1068   0.0996   0.0929   0.0866   0.0808   0.0753
#> 1  20.72749 20.72740 20.72731 20.72723 20.72715 20.72707 20.72699 20.72692
#> 2  21.07671 21.07666 21.07660 21.07655 21.07650 21.07645 21.07640 21.07636
#> 3  24.84355 24.84360 24.84365 24.84370 24.84374 24.84378 24.84381 24.84385
#> 4  21.64914 21.64906 21.64899 21.64892 21.64885 21.64879 21.64873 21.64867
#> 5  19.91570 19.91571 19.91572 19.91573 19.91574 19.91575 19.91576 19.91577
#> 6  14.36239 14.36217 14.36197 14.36177 14.36158 14.36139 14.36121 14.36104
#> 7  22.72265 22.72266 22.72267 22.72267 22.72268 22.72268 22.72269 22.72269
#> 8  20.95604 20.95617 20.95630 20.95643 20.95656 20.95668 20.95679 20.95691
#> 9  13.41783 13.41813 13.41841 13.41868 13.41894 13.41920 13.41944 13.41967
#> 10 17.41442 17.41454 17.41466 17.41477 17.41488 17.41499 17.41509 17.41519
#> 11 12.21741 12.21725 12.21711 12.21697 12.21683 12.21670 12.21657 12.21645
#> 12 11.73207 11.73205 11.73202 11.73199 11.73196 11.73194 11.73191 11.73189
#> 13 28.74105 28.74110 28.74115 28.74121 28.74126 28.74131 28.74135 28.74140
#> 14 33.05182 33.05174 33.05166 33.05159 33.05153 33.05147 33.05141 33.05136
#> 15 21.16935 21.16927 21.16918 21.16910 21.16902 21.16894 21.16885 21.16877
#> 16 13.41603 13.41607 13.41612 13.41616 13.41619 13.41623 13.41627 13.41630
#> 17 19.12621 19.12617 19.12615 19.12612 19.12610 19.12608 19.12606 19.12605
#> 18 28.43212 28.43214 28.43216 28.43218 28.43219 28.43221 28.43223 28.43224
#> 19 27.98446 27.98462 27.98477 27.98491 27.98505 27.98519 27.98532 27.98544
#> 20 28.33209 28.33200 28.33190 28.33181 28.33173 28.33164 28.33156 28.33149
#> 21 23.01146 23.01148 23.01150 23.01151 23.01153 23.01153 23.01154 23.01155
#>      0.0703   0.0655   0.0611   0.0570   0.0532   0.0496   0.0462   0.0431
#> 1  20.72684 20.72677 20.72670 20.72664 20.72657 20.72651 20.72645 20.72639
#> 2  21.07631 21.07627 21.07622 21.07618 21.07614 21.07610 21.07606 21.07603
#> 3  24.84388 24.84391 24.84393 24.84396 24.84398 24.84400 24.84402 24.84403
#> 4  21.64862 21.64857 21.64852 21.64848 21.64843 21.64839 21.64835 21.64832
#> 5  19.91578 19.91578 19.91579 19.91580 19.91580 19.91581 19.91581 19.91582
#> 6  14.36087 14.36071 14.36056 14.36041 14.36027 14.36013 14.36000 14.35988
#> 7  22.72269 22.72269 22.72269 22.72269 22.72269 22.72269 22.72269 22.72269
#> 8  20.95702 20.95712 20.95723 20.95733 20.95742 20.95752 20.95761 20.95769
#> 9  13.41989 13.42010 13.42030 13.42050 13.42068 13.42086 13.42103 13.42119
#> 10 17.41529 17.41539 17.41548 17.41557 17.41566 17.41575 17.41583 17.41591
#> 11 12.21634 12.21622 12.21612 12.21601 12.21591 12.21582 12.21573 12.21564
#> 12 11.73186 11.73184 11.73182 11.73180 11.73177 11.73175 11.73173 11.73171
#> 13 28.74145 28.74149 28.74153 28.74158 28.74162 28.74166 28.74170 28.74173
#> 14 33.05131 33.05127 33.05124 33.05120 33.05117 33.05115 33.05112 33.05110
#> 15 21.16869 21.16862 21.16854 21.16846 21.16839 21.16831 21.16824 21.16817
#> 16 13.41633 13.41636 13.41639 13.41641 13.41644 13.41646 13.41648 13.41650
#> 17 19.12604 19.12604 19.12603 19.12603 19.12603 19.12603 19.12603 19.12604
#> 18 28.43226 28.43227 28.43229 28.43230 28.43231 28.43233 28.43234 28.43235
#> 19 27.98556 27.98568 27.98579 27.98590 27.98601 27.98611 27.98621 27.98630
#> 20 28.33141 28.33134 28.33127 28.33121 28.33114 28.33108 28.33103 28.33097
#> 21 23.01155 23.01155 23.01155 23.01155 23.01154 23.01154 23.01153 23.01152
#>      0.0402   0.0375   0.0350   0.0326   0.0304   0.0284   0.0265   0.0247
#> 1  20.72633 20.72628 20.72622 20.72617 20.72612 20.72607 20.72603 20.52806
#> 2  21.07599 21.07596 21.07593 21.07589 21.07586 21.07583 21.07581 20.94426
#> 3  24.84405 24.84406 24.84407 24.84408 24.84409 24.84410 24.84411 24.88994
#> 4  21.64828 21.64825 21.64822 21.64819 21.64817 21.64814 21.64812 21.53628
#> 5  19.91582 19.91583 19.91583 19.91583 19.91584 19.91584 19.91584 19.90905
#> 6  14.35976 14.35964 14.35953 14.35943 14.35932 14.35923 14.35913 14.47960
#> 7  22.72269 22.72268 22.72268 22.72268 22.72267 22.72267 22.72266 22.61002
#> 8  20.95778 20.95786 20.95794 20.95801 20.95808 20.95815 20.95822 20.98333
#> 9  13.42135 13.42150 13.42164 13.42177 13.42190 13.42202 13.42214 13.43025
#> 10 17.41599 17.41606 17.41614 17.41621 17.41628 17.41634 17.41641 17.52571
#> 11 12.21556 12.21548 12.21540 12.21533 12.21526 12.21519 12.21513 12.23558
#> 12 11.73170 11.73168 11.73166 11.73164 11.73163 11.73161 11.73160 11.80927
#> 13 28.74177 28.74181 28.74184 28.74187 28.74190 28.74194 28.74196 28.77408
#> 14 33.05108 33.05107 33.05106 33.05105 33.05104 33.05103 33.05103 33.16324
#> 15 21.16810 21.16803 21.16796 21.16790 21.16783 21.16777 21.16771 21.16439
#> 16 13.41652 13.41654 13.41656 13.41658 13.41659 13.41661 13.41662 13.46323
#> 17 19.12604 19.12605 19.12606 19.12607 19.12608 19.12609 19.12610 18.95207
#> 18 28.43237 28.43238 28.43239 28.43240 28.43241 28.43242 28.43244 28.39693
#> 19 27.98639 27.98648 27.98656 27.98664 27.98672 27.98679 27.98686 27.91154
#> 20 28.33092 28.33087 28.33082 28.33077 28.33073 28.33069 28.33065 28.49015
#> 21 23.01151 23.01150 23.01149 23.01148 23.01147 23.01146 23.01145 23.10301
#>      0.0230   0.0215   0.0200   0.0187   0.0174   0.0162   0.0151   0.0141
#> 1  20.52791 20.52778 20.52955 20.52930 20.52907 20.52885 20.52863 20.52843
#> 2  20.94418 20.94410 20.97464 20.97455 20.97445 20.97436 20.97428 20.97420
#> 3  24.88993 24.88992 24.84717 24.84711 24.84705 24.84699 24.84693 24.84688
#> 4  21.53623 21.53619 21.47127 21.47102 21.47079 21.47057 21.47037 21.47017
#> 5  19.90906 19.90907 19.89984 19.89983 19.89983 19.89982 19.89981 19.89980
#> 6  14.47926 14.47893 14.49736 14.49749 14.49761 14.49772 14.49782 14.49792
#> 7  22.61001 22.61000 22.64194 22.64190 22.64186 22.64182 22.64179 22.64175
#> 8  20.98357 20.98381 20.87918 20.87914 20.87910 20.87907 20.87904 20.87901
#> 9  13.43063 13.43099 13.45135 13.45139 13.45143 13.45147 13.45150 13.45154
#> 10 17.52594 17.52616 17.61427 17.61459 17.61489 17.61518 17.61546 17.61572
#> 11 12.23536 12.23515 12.25426 12.25433 12.25439 12.25445 12.25450 12.25455
#> 12 11.80920 11.80913 11.81328 11.81338 11.81347 11.81355 11.81363 11.81370
#> 13 28.77419 28.77430 28.76294 28.76297 28.76300 28.76302 28.76305 28.76307
#> 14 33.16325 33.16328 33.20379 33.20404 33.20428 33.20450 33.20472 33.20493
#> 15 21.16413 21.16387 21.23569 21.23571 21.23573 21.23575 21.23576 21.23577
#> 16 13.46325 13.46327 13.40734 13.40724 13.40714 13.40705 13.40697 13.40688
#> 17 18.95219 18.95231 18.93407 18.93384 18.93362 18.93342 18.93323 18.93304
#> 18 28.39698 28.39702 28.35672 28.35660 28.35648 28.35637 28.35627 28.35617
#> 19 27.91181 27.91206 28.02246 28.02258 28.02268 28.02279 28.02289 28.02298
#> 20 28.49000 28.48985 28.45218 28.45234 28.45250 28.45264 28.45278 28.45291
#> 21 23.10291 23.10281 23.05069 23.05066 23.05063 23.05061 23.05058 23.05056
#>      0.0132   0.0123   0.0115   0.0107   0.0100   0.0093   0.0087   0.0081
#> 1  20.52823 20.52805 20.52787 20.52770 20.52753 20.52738 20.52723 20.52709
#> 2  20.97412 20.97404 20.97397 20.97389 20.97383 20.97376 20.97370 20.97364
#> 3  24.84683 24.84678 24.84673 24.84669 24.84664 24.84660 24.84656 24.84653
#> 4  21.46999 21.46981 21.46965 21.46949 21.46934 21.46920 21.46907 21.46894
#> 5  19.89979 19.89979 19.89978 19.89977 19.89977 19.89976 19.89976 19.89975
#> 6  14.49801 14.49810 14.49817 14.49824 14.49831 14.49837 14.49843 14.49848
#> 7  22.64172 22.64169 22.64165 22.64162 22.64159 22.64156 22.64153 22.64151
#> 8  20.87899 20.87898 20.87896 20.87895 20.87895 20.87894 20.87894 20.87894
#> 9  13.45158 13.45162 13.45165 13.45169 13.45172 13.45176 13.45179 13.45182
#> 10 17.61597 17.61621 17.61643 17.61665 17.61686 17.61705 17.61724 17.61742
#> 11 12.25460 12.25464 12.25468 12.25472 12.25475 12.25478 12.25481 12.25483
#> 12 11.81377 11.81384 11.81390 11.81396 11.81401 11.81406 11.81411 11.81416
#> 13 28.76310 28.76312 28.76315 28.76317 28.76320 28.76322 28.76325 28.76327
#> 14 33.20513 33.20532 33.20550 33.20567 33.20584 33.20599 33.20614 33.20629
#> 15 21.23577 21.23577 21.23577 21.23576 21.23575 21.23574 21.23573 21.23571
#> 16 13.40681 13.40673 13.40666 13.40660 13.40654 13.40648 13.40643 13.40638
#> 17 18.93287 18.93271 18.93256 18.93242 18.93229 18.93217 18.93205 18.93194
#> 18 28.35608 28.35599 28.35591 28.35584 28.35576 28.35570 28.35563 28.35557
#> 19 28.02307 28.02316 28.02324 28.02332 28.02340 28.02347 28.02354 28.02360
#> 20 28.45304 28.45316 28.45327 28.45337 28.45347 28.45357 28.45366 28.45374
#> 21 23.05053 23.05051 23.05049 23.05046 23.05044 23.05042 23.05040 23.05038
#>      0.0075   0.0070   0.0066   0.0061   0.0057   0.0053
#> 1  20.52695 20.52683 20.52670 20.52659 20.52647 20.52637
#> 2  20.97358 20.97352 20.97347 20.97342 20.97337 20.97332
#> 3  24.84649 24.84646 24.84643 24.84640 24.84637 24.84634
#> 4  21.46883 21.46871 21.46861 21.46851 21.46842 21.46833
#> 5  19.89975 19.89975 19.89974 19.89974 19.89973 19.89973
#> 6  14.49853 14.49857 14.49862 14.49865 14.49869 14.49872
#> 7  22.64148 22.64145 22.64143 22.64140 22.64138 22.64135
#> 8  20.87894 20.87894 20.87895 20.87896 20.87896 20.87897
#> 9  13.45186 13.45189 13.45192 13.45195 13.45197 13.45200
#> 10 17.61759 17.61775 17.61790 17.61805 17.61819 17.61832
#> 11 12.25486 12.25488 12.25490 12.25491 12.25493 12.25495
#> 12 11.81420 11.81424 11.81428 11.81432 11.81435 11.81438
#> 13 28.76329 28.76331 28.76333 28.76335 28.76338 28.76340
#> 14 33.20642 33.20655 33.20668 33.20680 33.20691 33.20702
#> 15 21.23570 21.23568 21.23566 21.23564 21.23562 21.23560
#> 16 13.40633 13.40628 13.40624 13.40620 13.40616 13.40613
#> 17 18.93184 18.93175 18.93166 18.93158 18.93150 18.93143
#> 18 28.35552 28.35547 28.35542 28.35537 28.35533 28.35529
#> 19 28.02367 28.02373 28.02378 28.02383 28.02389 28.02393
#> 20 28.45382 28.45390 28.45397 28.45404 28.45410 28.45417
#> 21 23.05036 23.05034 23.05032 23.05030 23.05028 23.05026
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 22.8 21.4 18.1 14.3 24.4 19.2 16.4 15.2 10.4 14.7 32.4 33.9 21.5
#> [16] 13.3 19.2 27.3 26.0 30.4 21.4
#> 
#> $X
#>               am       carb        cyl        disp       drat       gear
#>  [1,]  1.1547005  1.2508618  0.1097643 -0.43123068  0.6087775  0.5834600
#>  [2,]  1.1547005  1.2508618  0.1097643 -0.43123068  0.6087775  0.5834600
#>  [3,]  1.1547005 -1.1371471 -1.0427605 -0.85462925  0.5025954  0.5834600
#>  [4,] -0.8660254 -1.1371471  0.1097643  0.36671280 -1.1326093 -0.9481224
#>  [5,] -0.8660254 -1.1371471  0.1097643  0.09801755 -1.8121749 -0.9481224
#>  [6,] -0.8660254  1.2508618  1.2622890  1.19722539 -0.8565358 -0.9481224
#>  [7,] -0.8660254 -0.3411441 -1.0427605 -0.53952300  0.1628126  0.5834600
#>  [8,] -0.8660254  1.2508618  0.1097643 -0.36934934  0.6512504  0.5834600
#>  [9,] -0.8660254  0.4548588  1.2622890  0.51164539 -1.1538457 -0.9481224
#> [10,] -0.8660254  0.4548588  1.2622890  0.51164539 -1.1538457 -0.9481224
#> [11,] -0.8660254  1.2508618  1.2622890  2.01145342 -1.3025007 -0.9481224
#> [12,] -0.8660254  1.2508618  1.2622890  1.84860782 -0.8140629 -0.9481224
#> [13,]  1.1547005 -1.1371471 -1.0427605 -1.09319807  0.9910331  0.5834600
#> [14,]  1.1547005 -1.1371471 -1.0427605 -1.15507940  1.2883431  0.5834600
#> [15,] -0.8660254 -1.1371471 -1.0427605 -0.75610766  0.1840490 -0.9481224
#> [16,] -0.8660254  1.2508618  1.2622890  1.11580259  0.2477583 -0.9481224
#> [17,] -0.8660254 -0.3411441  1.2622890  1.52291660 -1.1326093 -0.9481224
#> [18,]  1.1547005 -1.1371471 -1.0427605 -1.09075538  0.9910331  0.5834600
#> [19,]  1.1547005 -0.3411441 -1.0427605 -0.75447920  1.7343080  2.1150424
#> [20,]  1.1547005 -0.3411441 -1.0427605 -0.95966467  0.3327040  2.1150424
#> [21,]  1.1547005 -0.3411441 -1.0427605 -0.74877961  1.0547424  0.5834600
#>               hp        qsec         vs          wt
#>  [1,] -0.3846524 -1.14418818 -1.0488088 -0.50430810
#>  [2,] -0.3846524 -0.74620969 -1.0488088 -0.25522400
#>  [3,] -0.6707377  0.38376498  0.9534626 -0.79734823
#>  [4,] -0.3846524  0.97362597  0.9534626  0.07688815
#>  [5,] -0.4687952  1.52795317  0.9534626  0.31620425
#>  [6,]  1.8872010 -1.58480724 -1.0488088  0.42365230
#>  [7,] -1.1924225  1.37160447  0.9534626  0.05246814
#>  [8,] -0.1658814  0.16345545  0.9534626  0.29666824
#>  [9,]  0.7933456 -0.47615285 -1.0488088  0.91205251
#> [10,]  0.7933456 -0.04974731 -1.0488088  0.62878039
#> [11,]  1.3823447 -0.17766897 -1.0488088  2.23464028
#> [12,]  1.6347728 -0.46193933 -1.0488088  2.15747305
#> [13,] -1.1251084  0.99494625  0.9534626 -0.91456428
#> [14,] -1.1419369  1.30053688  0.9534626 -1.27109644
#> [15,] -0.6034235  1.37871123  0.9534626 -0.65571217
#> [16,]  1.8872010 -1.89039787 -1.0488088  0.68738841
#> [17,]  0.7092029 -0.72488941 -1.0488088  0.69227241
#> [18,] -1.1251084  0.58986099  0.9534626 -1.17341639
#> [19,] -0.7043948 -0.97362597 -1.0488088 -0.97317231
#> [20,] -0.3341668 -0.83149079  0.9534626 -1.58562617
#> [21,] -0.4014810  0.37665822  0.9534626 -0.34802004
#> attr(,"center")
#>  [1]   0.4285714   2.4285714   5.8095238 212.9619048   3.6133333   3.6190476
#>  [7] 132.8571429  18.0700000   0.5238095   3.1362857
#> attr(,"scale")
#>  [1]   0.4948717   1.2562768   1.7353207 122.8157175   0.4708891   0.6529195
#>  [7]  59.4228477   1.4071112   0.4994328   1.0237506
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 35
#> 
#> $lambda.min
#> [1] 0.4957438
#> 
#> $null.dev
#> [1] 37.62816
#> 
#> $Bias
#> [1] 11.92222
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#>  11.8304