Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error', predict_raw = 'FALSE'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 30.001152 28.072817 25.937417 23.606193 21.305217 19.380407 17.777614
#>   [8] 16.450088 15.357494 14.465052 13.742793 13.164916 12.709224 12.356640
#>  [15] 12.090785 11.878662 11.585603 11.371340 11.198826 11.100871 10.968571
#>  [22] 10.856799 10.734515 10.577913 10.446539 10.327665 10.228108 10.050373
#>  [29]  9.876193  9.728912  9.604791  9.115764  9.041221  8.697396  8.599972
#>  [36]  8.207141  8.239664  8.704473 12.205034 11.732652 12.147316 12.368504
#>  [43] 14.585707 15.421711 15.900629 16.283110 16.935899 16.936425 16.653059
#>  [50] 16.653420 16.503056 16.503502 16.519159 16.520056 16.336699 16.336868
#>  [57] 18.127230 17.859195 17.859646 17.789899 17.414890 17.414683 17.187513
#>  [64] 17.187366 17.367369 19.401726 19.317498 19.143292 19.144456 19.145575
#>  [71] 19.146651 19.416592 19.417918 19.781089 19.340354 19.341625 19.342853
#>  [78] 19.420176 19.421486 19.380965 19.382013 19.383022 19.383994 19.384931
#>  [85] 19.385833 19.386702 19.387540 19.388347 19.389125 19.389874 19.390595
#>  [92] 19.391291 19.391960 19.392605 19.393227 19.393826 19.394402 19.394958
#>  [99] 19.395493 19.396008
#> 
#> $cvse
#>   [1] 9.939836 9.866776 9.671412 9.152112 8.442757 7.808181 7.239498 6.729165
#>   [9] 6.270767 5.858841 5.488710 5.156348 4.858269 4.591419 4.353100 4.141827
#>  [17] 3.959091 3.795617 3.652346 3.624791 3.630360 3.639261 3.641754 3.634266
#>  [25] 3.629966 3.625659 3.624293 3.571244 3.515560 3.465506 3.420540 3.407335
#>  [33] 3.374111 3.303858 3.260220 2.903091 2.898815 2.944589 4.418211 4.421220
#>  [41] 4.423742 4.468606 4.823343 5.109132 5.359930 5.592989 5.639030 5.638711
#>  [49] 5.623279 5.623059 5.542006 5.541900 5.349255 5.348780 5.358978 5.358748
#>  [57] 5.524197 5.537227 5.537155 5.545393 5.728209 5.728071 5.735205 5.735076
#>  [65] 5.918275 7.554369 7.563008 7.578339 7.579454 7.580527 7.581559 7.570505
#>  [73] 7.571444 7.678676 7.684808 7.685775 7.686705 7.682584 7.683433 7.688396
#>  [81] 7.689196 7.689966 7.690707 7.691420 7.692106 7.692766 7.693401 7.694011
#>  [89] 7.694598 7.695163 7.695706 7.696229 7.696731 7.697214 7.697679 7.698126
#>  [97] 7.698556 7.698969 7.699367 7.699749
#> 
#> $fold
#>  [1]  4 10  5  9  7  2  8  1  4  2  3  6  1  9  5  7  3  8 10  1  6
#> 
#> $lambda
#>   [1] 4.718264835 4.400269577 4.103706134 3.827130076 3.569194317 3.328642566
#>   [7] 3.104303198 2.895083552 2.699964610 2.517996032 2.348291527 2.190024537
#>  [13] 2.042424213 1.904771657 1.776396422 1.656673249 1.545019017 1.440889906
#>  [19] 1.343778749 1.253212559 1.168750226 1.089980373 1.016519344 0.948009342
#>  [25] 0.884116685 0.824530180 0.768959605 0.717134302 0.668801850 0.623726843
#>  [31] 0.581689742 0.542485800 0.505924073 0.471826484 0.440026958 0.410370613
#>  [37] 0.382713008 0.356919432 0.332864257 0.310430320 0.289508355 0.269996461
#>  [43] 0.251799603 0.234829153 0.219002454 0.204242421 0.190477166 0.177639642
#>  [49] 0.165667325 0.154501902 0.144088991 0.134377875 0.125321256 0.116875023
#>  [55] 0.108998037 0.101651934 0.094800934 0.088411669 0.082453018 0.076895961
#>  [61] 0.071713430 0.066880185 0.062372684 0.058168974 0.054248580 0.050592407
#>  [67] 0.047182648 0.044002696 0.041037061 0.038271301 0.035691943 0.033286426
#>  [73] 0.031043032 0.028950836 0.026999646 0.025179960 0.023482915 0.021900245
#>  [79] 0.020424242 0.019047717 0.017763964 0.016566732 0.015450190 0.014408899
#>  [85] 0.013437787 0.012532126 0.011687502 0.010899804 0.010165193 0.009480093
#>  [91] 0.008841167 0.008245302 0.007689596 0.007171343 0.006688018 0.006237268
#>  [97] 0.005816897 0.005424858 0.005059241 0.004718265
#> 
#> $fit
#> $beta
#>               4.7183     4.4003    4.1037    3.8271    3.5692    3.3286
#> (Intercept) 20.40476 22.4145939 24.288970 26.037019 27.667256 29.187620
#> am           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000 -0.6587556 -1.273113 -1.846065 -2.380402 -2.878727
#>                3.1043    2.8951    2.7000    2.5180    2.3483    2.1900
#> (Intercept) 30.605517 31.927852 33.161066 34.311166 35.383753 36.384051
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -3.343466 -3.776883 -4.181089 -4.558054 -4.909612 -5.237476
#>                2.0424    1.9048    1.7764    1.6567    1.5450    1.4409
#> (Intercept) 37.316932 38.186941 38.998313 39.755002 40.285393 40.285393
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -5.543243 -5.828403 -6.094343 -6.342361 -6.516205 -6.516205
#>                1.3438      1.2532     1.1688    1.0900     1.0165     0.9480
#> (Intercept) 40.285393 39.63320592 37.7922322 36.075581 34.4746258 32.9826045
#> am           0.000000  0.00000000  0.0000000  0.000000  0.0000000  0.0000000
#> carb         0.000000  0.00000000  0.0000000  0.000000  0.0000000  0.0000000
#> cyl          0.000000  0.00000000  0.0000000  0.000000  0.0000000  0.0000000
#> disp         0.000000  0.00000000  0.0000000  0.000000  0.0000000  0.0000000
#> drat         0.000000  0.00000000  0.0000000  0.000000  0.0000000  0.0000000
#> gear         0.000000  0.00000000  0.0000000  0.000000  0.0000000  0.0000000
#> hp           0.000000  0.00000000  0.0000000  0.000000  0.0000000  0.0000000
#> qsec         0.000000  0.03236401  0.1237202  0.208907  0.2883525  0.3623923
#> vs           0.000000  0.00000000  0.0000000  0.000000  0.0000000  0.0000000
#> wt          -6.516205 -6.49106243 -6.4200910 -6.353912 -6.2921938 -6.2346748
#>                 0.8841     0.8245    0.7690    0.7171     0.6688    0.6237
#> (Intercept) 31.5901067 30.2914578 29.080333 27.950835 26.8974607 25.915080
#> am           0.0000000  0.0000000  0.000000  0.000000  0.0000000  0.000000
#> carb         0.0000000  0.0000000  0.000000  0.000000  0.0000000  0.000000
#> cyl          0.0000000  0.0000000  0.000000  0.000000  0.0000000  0.000000
#> disp         0.0000000  0.0000000  0.000000  0.000000  0.0000000  0.000000
#> drat         0.0000000  0.0000000  0.000000  0.000000  0.0000000  0.000000
#> gear         0.0000000  0.0000000  0.000000  0.000000  0.0000000  0.000000
#> hp           0.0000000  0.0000000  0.000000  0.000000  0.0000000  0.000000
#> qsec         0.4314934  0.4959373  0.556038  0.612088  0.6643605  0.713110
#> vs           0.0000000  0.0000000  0.000000  0.000000  0.0000000  0.000000
#> wt          -6.1809926 -6.1309284 -6.084238 -6.040695 -6.0000862 -5.962214
#>                 0.5817     0.5425     0.5059     0.4718     0.4400     0.4104
#> (Intercept) 24.9989093 24.1444850 23.3476460 22.8346801 22.8340174 22.8339423
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.7585739  0.8009737  0.8405159  0.8659712  0.8660041  0.8660078
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.9268951 -5.8939562 -5.8632372 -5.8434618 -5.8434363 -5.8434334
#>                 0.3827     0.3569     0.3329     0.3104      0.2895     0.2700
#> (Intercept) 22.8339338 22.8339328 22.8339327 22.8339327 22.30277038 20.2933624
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.00000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.00000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.00000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.00000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.00000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.05259689  0.2509881
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.00000000  0.0000000
#> qsec         0.8660083  0.8660083  0.8660083  0.8660083  0.87909094  0.9287306
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.00000000  0.0000000
#> wt          -5.8434331 -5.8434330 -5.8434330 -5.8434330 -5.80879484 -5.6779147
#>                 0.2518     0.2348       0.2190       0.2042       0.1905
#> (Intercept) 18.4166302  0.9501779 -60.26173206 -60.26971454 -60.27718370
#> am           0.0000000  0.0000000   5.07540368   5.07596106   5.07648259
#> carb         0.0000000  0.0000000   0.00000000   0.00000000   0.00000000
#> cyl          0.0000000  0.0000000   0.00000000   0.00000000   0.00000000
#> disp         0.0000000  0.0000000   0.02650047   0.02650373   0.02650678
#> drat         0.0000000  0.0000000   4.43745373   4.43783223   4.43818639
#> gear         0.4362686  1.4035723   0.00000000   0.00000000   0.00000000
#> hp           0.0000000  0.0000000   0.00000000   0.00000000   0.00000000
#> qsec         0.9750956  1.8006066   3.82738773   3.82767437   3.82794257
#> vs           0.0000000 -2.6682927  -4.49760339  -4.49791103  -4.49819888
#> wt          -5.5556793 -5.4296656  -3.10113556  -3.10090609  -3.10069138
#>                   0.1776       0.1657       0.1545       0.1441       0.1344
#> (Intercept) -60.28417254 -60.29071194 -60.29683081 -60.30255619 -60.30791339
#> am            5.07697059   5.07742720   5.07785445   5.07825423   5.07862829
#> carb          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> cyl           0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> disp          0.02650964   0.02651232   0.02651482   0.02651716   0.02651935
#> drat          4.43851777   4.43882784   4.43911797   4.43938944   4.43964345
#> gear          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> hp            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> qsec          3.82819353   3.82842835   3.82864807   3.82885366   3.82904602
#> vs           -4.49846823  -4.49872025  -4.49895606  -4.49917671  -4.49938318
#> wt           -3.10049049  -3.10030251  -3.10012662  -3.09996204  -3.09980804
#>                  0.1253       0.1169       0.1090       0.1017       0.0948
#> (Intercept) -60.3129261 -60.31761643 -61.19134003 -61.19561831 -61.19972291
#> am            5.0789783   5.07930581   4.98572678   4.98587289   4.98601341
#> carb          0.0000000   0.00000000   0.44292832   0.44306076   0.44318757
#> cyl           0.0000000   0.00000000  -1.07101362  -1.07119728  -1.07137318
#> disp          0.0265214   0.02652332   0.04143833   0.04144343   0.04144831
#> drat          4.4398811   4.44010352   3.98960301   3.98964848   3.98969246
#> gear          0.0000000   0.00000000   0.00000000   0.00000000   0.00000000
#> hp            0.0000000   0.00000000   0.00000000   0.00000000   0.00000000
#> qsec          3.8292260   3.82939444   4.20942692   4.20968631   4.20993495
#> vs           -4.4995764  -4.49975712  -5.79818380  -5.79866317  -5.79912245
#> wt           -3.0996640  -3.09952913  -3.56172330  -3.56194238  -3.56215170
#>                   0.0884       0.0825       0.0769       0.0717       0.0669
#> (Intercept) -61.20366064 -61.20743807 -61.21106153 -61.21453713 -57.11286008
#> am            4.98614855   4.98627847   4.98640336   4.98652341   5.02723320
#> carb          0.44330898   0.44342522   0.44353653   0.44364311   0.62945954
#> cyl          -1.07154166  -1.07170304  -1.07185761  -1.07200566  -1.02211288
#> disp          0.04145299   0.04145747   0.04146176   0.04146588   0.04477647
#> drat          3.98973497   3.98977604   3.98981571   3.98985400   3.80301342
#> gear          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> hp            0.00000000   0.00000000   0.00000000   0.00000000  -0.01037320
#> qsec          4.21017327   4.21040170   4.21062064   4.21083049   4.00986025
#> vs           -5.79956248  -5.79998407  -5.80038799  -5.80077499  -5.27498092
#> wt           -3.56235172  -3.56254286  -3.56272556  -3.56290018  -3.61251591
#>                   0.0624       0.0582       0.0542       0.0506       0.0472
#> (Intercept) -57.11605320 -57.11923566 -57.12239833 -57.12553316 -57.12863303
#> am            5.02717519   5.02712934   5.02709450   5.02706961   5.02705370
#> carb          0.62960240   0.62973634   0.62986193   0.62997970   0.63009015
#> cyl          -1.02236342  -1.02260148  -1.02282768  -1.02304260  -1.02324680
#> disp          0.04478143   0.04478614   0.04479062   0.04479488   0.04479893
#> drat          3.80293708   3.80287377   3.80282224   3.80278134   3.80275001
#> gear          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> hp           -0.01037118  -0.01036921  -0.01036730  -0.01036544  -0.01036363
#> qsec          4.01013604   4.01040269   4.01066035   4.01090916   4.01114931
#> vs           -5.27562957  -5.27625249  -5.27685046  -5.27742425  -5.27797465
#> wt           -3.61294538  -3.61334369  -3.61371302  -3.61405539  -3.61437267
#>                   0.0440       0.0410       0.0383       0.0357       0.0333
#> (Intercept) -57.13169170 -57.13470377 -57.13766456 -57.14057006 -57.14341691
#> am            5.02704585   5.02704522   5.02705103   5.02706257   5.02707918
#> carb          0.63019376   0.63029095   0.63038213   0.63046769   0.63054799
#> cyl          -1.02344080  -1.02362511  -1.02380021  -1.02396658  -1.02412463
#> disp          0.04480278   0.04480643   0.04480991   0.04481322   0.04481637
#> drat          3.80272727   3.80271224   3.80270409   3.80270206   3.80270547
#> gear          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> hp           -0.01036188  -0.01036018  -0.01035854  -0.01035695  -0.01035542
#> qsec          4.01138097   4.01160433   4.01181958   4.01202693   4.01222658
#> vs           -5.27850242  -5.27900834  -5.27949316  -5.27995763  -5.28040247
#> wt           -3.61466664  -3.61493895  -3.61519112  -3.61542459  -3.61564070
#>                   0.0310       0.0290       0.0270        0.0252       0.0235
#> (Intercept) -57.14620229 -57.14892390 -57.15157990 -5.715239e+01 -56.73101736
#> am            5.02710026   5.02712524   5.02715363  5.027185e+00   5.27205291
#> carb          0.63062337   0.63069412   0.63076056  6.308230e-01   0.66030198
#> cyl          -1.02427480  -1.02441748  -1.02455303 -1.024682e+00  -1.11256056
#> disp          0.04481936   0.04482221   0.04482491  4.482749e-02   0.04550304
#> drat          3.80271368   3.80272613   3.80274228  3.802762e+00   3.88197573
#> gear          0.00000000   0.00000000   0.00000000 -2.561869e-04  -0.23236953
#> hp           -0.01035394  -0.01035251  -0.01035114 -1.034957e-02  -0.01015709
#> qsec          4.01241874   4.01260361   4.01278140  4.012921e+00   4.01122515
#> vs           -5.28082843  -5.28123619  -5.28162645 -5.281889e+00  -5.28679563
#> wt           -3.61584069  -3.61602572  -3.61619687 -3.616474e+00  -3.49734183
#>                   0.0219       0.0204       0.0190       0.0178       0.0166
#> (Intercept) -56.73641615 -56.74154757 -56.74642249 -56.75105166 -56.75544567
#> am            5.27276652   5.27343945   5.27407390   5.27467195   5.27523560
#> carb          0.66028680   0.66027273   0.66025975   0.66024782   0.66023692
#> cyl          -1.11271102  -1.11285196  -1.11298409  -1.11310808  -1.11322451
#> disp          0.04550479   0.04550647   0.04550808   0.04550962   0.04551109
#> drat          3.88242590   3.88285055   3.88325100   3.88362850   3.88398428
#> gear         -0.23256324  -0.23274438  -0.23291386  -0.23307249  -0.23322103
#> hp           -0.01015697  -0.01015684  -0.01015672  -0.01015660  -0.01015648
#> qsec          4.01137842   4.01152483   4.01166467   4.01179819   4.01192564
#> vs           -5.28699631  -5.28718706  -5.28736848  -5.28754115  -5.28770557
#> wt           -3.49665186  -3.49600482  -3.49539820  -3.49482963  -3.49429685
#>                   0.0155       0.0144       0.0134       0.0125       0.0117
#> (Intercept) -56.75961491 -56.76356952 -56.76731939 -56.77087412 -56.77424303
#> am            5.27576676   5.27626723   5.27673875   5.27718294   5.27760137
#> carb          0.66022700   0.66021801   0.66020993   0.66020270   0.66019627
#> cyl          -1.11333396  -1.11343693  -1.11353388  -1.11362524  -1.11371141
#> disp          0.04551250   0.04551384   0.04551513   0.04551636   0.04551755
#> drat          3.88431950   3.88463529   3.88493270   3.88521275   3.88547643
#> gear         -0.23336020  -0.23349063  -0.23361294  -0.23372769  -0.23383540
#> hp           -0.01015636  -0.01015624  -0.01015612  -0.01015600  -0.01015588
#> qsec          4.01204727   4.01216333   4.01227405   4.01237967   4.01248040
#> vs           -5.28786224  -5.28801159  -5.28815406  -5.28829003  -5.28841985
#> wt           -3.49379771  -3.49333020  -3.49289239  -3.49248247  -3.49209871
#>                   0.0109       0.0102       0.0095       0.0088       0.0082
#> (Intercept) -56.77743510 -56.78045903 -56.78332318 -56.78603558 -56.78860396
#> am            5.27799549   5.27836671   5.27871635   5.27904565   5.27935581
#> carb          0.66019060   0.66018565   0.66018136   0.66017771   0.66017463
#> cyl          -1.11379274  -1.11386958  -1.11394222  -1.11401094  -1.11407600
#> disp          0.04551868   0.04551976   0.04552079   0.04552179   0.04552274
#> drat          3.88572465   3.88595829   3.88617819   3.88638514   3.88657990
#> gear         -0.23393654  -0.23403157  -0.23412090  -0.23420490  -0.23428394
#> hp           -0.01015576  -0.01015564  -0.01015552  -0.01015540  -0.01015528
#> qsec          4.01257647   4.01266807   4.01275541   4.01283869   4.01291808
#> vs           -5.28854387  -5.28866240  -5.28877572  -5.28888412  -5.28898784
#> wt           -3.49173949  -3.49140329  -3.49108867  -3.49079425  -3.49051877
#>                   0.0077       0.0072       0.0067       0.0062       0.0058
#> (Intercept) -56.79103570 -56.79333788 -56.79551727 -56.79758032 -56.79953319
#> am            5.27964794   5.27992310   5.28018228   5.28042643   5.28065644
#> carb          0.66017210   0.66017007   0.66016849   0.66016735   0.66016659
#> cyl          -1.11413765  -1.11419610  -1.11425155  -1.11430419  -1.11435419
#> disp          0.04552365   0.04552452   0.04552535   0.04552615   0.04552691
#> drat          3.88676317   3.88693564   3.88709793   3.88725065   3.88739437
#> gear         -0.23435835  -0.23442842  -0.23449445  -0.23455669  -0.23461540
#> hp           -0.01015516  -0.01015505  -0.01015493  -0.01015481  -0.01015470
#> qsec          4.01299377   4.01306593   4.01313471   4.01320028   4.01326280
#> vs           -5.28908712  -5.28918218  -5.28927323  -5.28936046  -5.28944406
#> wt           -3.49026102  -3.49001986  -3.48979424  -3.48958315  -3.48938566
#>                   0.0054       0.0051       0.0047
#> (Intercept) -56.80138175 -56.80313157 -56.80478798
#> am            5.28087313   5.28107731   5.28126972
#> carb          0.66016619   0.66016611   0.66016633
#> cyl          -1.11440172  -1.11444691  -1.11448991
#> disp          0.04552765   0.04552835   0.04552902
#> drat          3.88752963   3.88765694   3.88777676
#> gear         -0.23467079  -0.23472309  -0.23477247
#> hp           -0.01015458  -0.01015447  -0.01015435
#> qsec          4.01332239   4.01337921   4.01343337
#> vs           -5.28952420  -5.28960103  -5.28967472
#> wt           -3.48920088  -3.48902798  -3.48886621
#> 
#> $iter
#>   [1]   0   5   3   3   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   5   5   5   5   4   4   4   4   4   4   4   4   4   4   3   1   1
#>  [37]   1   1   1   1  16  22  22  73 202   1   1   1   1   1   1   1   1   1
#>  [55] 182   1   1   1   1   1   1  87   1   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1  62   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.718264835 4.400269577 4.103706134 3.827130076 3.569194317 3.328642566
#>   [7] 3.104303198 2.895083552 2.699964610 2.517996032 2.348291527 2.190024537
#>  [13] 2.042424213 1.904771657 1.776396422 1.656673249 1.545019017 1.440889906
#>  [19] 1.343778749 1.253212559 1.168750226 1.089980373 1.016519344 0.948009342
#>  [25] 0.884116685 0.824530180 0.768959605 0.717134302 0.668801850 0.623726843
#>  [31] 0.581689742 0.542485800 0.505924073 0.471826484 0.440026958 0.410370613
#>  [37] 0.382713008 0.356919432 0.332864257 0.310430320 0.289508355 0.269996461
#>  [43] 0.251799603 0.234829153 0.219002454 0.204242421 0.190477166 0.177639642
#>  [49] 0.165667325 0.154501902 0.144088991 0.134377875 0.125321256 0.116875023
#>  [55] 0.108998037 0.101651934 0.094800934 0.088411669 0.082453018 0.076895961
#>  [61] 0.071713430 0.066880185 0.062372684 0.058168974 0.054248580 0.050592407
#>  [67] 0.047182648 0.044002696 0.041037061 0.038271301 0.035691943 0.033286426
#>  [73] 0.031043032 0.028950836 0.026999646 0.025179960 0.023482915 0.021900245
#>  [79] 0.020424242 0.019047717 0.017763964 0.016566732 0.015450190 0.014408899
#>  [85] 0.013437787 0.012532126 0.011687502 0.010899804 0.010165193 0.009480093
#>  [91] 0.008841167 0.008245302 0.007689596 0.007171343 0.006688018 0.006237268
#>  [97] 0.005816897 0.005424858 0.005059241 0.004718265
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 40
#> 
#> $loss
#>   [1] 603.82952 514.08318 438.99702 376.46133 324.65468 282.00547 247.15853
#>   [8] 218.94617 196.36297 178.54400 164.74588 154.33032 146.74988 141.53565
#>  [15] 138.28649 136.65979 136.32704 136.32704 136.32704 133.46687 125.98094
#>  [22] 119.78252 114.68193 110.51928 107.14859 104.45257 102.32744 100.68405
#>  [29]  99.44584  98.54714  97.93173  97.55150  97.36537  97.33158  97.33158
#>  [36]  97.33158  97.33158  97.33158  97.33158  97.33158  96.88100  95.43179
#>  [43]  94.43572  86.18717  60.61888  60.61887  60.61886  60.61885  60.61885
#>  [50]  60.61884  60.61884  60.61883  60.61883  60.61882  55.29316  55.29315
#>  [57]  55.29313  55.29312  55.29311  55.29310  55.29309  54.23958  54.23956
#>  [64]  54.23955  54.23953  54.23951  54.23950  54.23949  54.23948  54.23947
#>  [71]  54.23946  54.23945  54.23944  54.23944  54.23943  54.23923  54.14954
#>  [78]  54.14952  54.14951  54.14950  54.14949  54.14948  54.14947  54.14947
#>  [85]  54.14946  54.14946  54.14945  54.14945  54.14944  54.14944  54.14944
#>  [92]  54.14944  54.14943  54.14943  54.14943  54.14943  54.14943  54.14943
#>  [99]  54.14942  54.14942
#> 
#> $linear.predictors
#>      4.7183   4.4003   4.1037   3.8271   3.5692   3.3286   3.1043   2.8951
#> 1  20.40476 20.68865 20.95341 21.20033 21.43060 21.64536 21.84564 22.03242
#> 2  20.40476 20.52067 20.62877 20.72958 20.82360 20.91128 20.99305 21.06931
#> 3  20.40476 20.88628 21.33535 21.75415 22.14472 22.50897 22.84868 23.16548
#> 4  20.40476 20.29669 20.19591 20.10192 20.01426 19.93251 19.85627 19.78517
#> 5  20.40476 20.13530 19.88400 19.64963 19.43106 19.22723 19.03713 18.85984
#> 6  20.40476 20.31316 20.22774 20.14807 20.07377 20.00448 19.93986 19.87960
#> 7  20.40476 20.14847 19.90946 19.68655 19.47867 19.28480 19.10399 18.93537
#> 8  20.40476 19.73346 19.10740 18.52353 17.97902 17.47120 16.99761 16.55594
#> 9  20.40476 19.95744 19.54026 19.15120 18.78836 18.44997 18.13439 17.84008
#> 10 20.40476 19.92450 19.47660 19.05889 18.66934 18.30603 17.96722 17.65123
#> 11 20.40476 21.35070 22.23289 23.05562 23.82291 24.53848 25.20582 25.82819
#> 12 20.40476 21.20578 21.95281 22.64949 23.29922 23.90516 24.47026 24.99727
#> 13 20.40476 20.79076 21.15075 21.48647 21.79956 22.09156 22.36387 22.61784
#> 14 20.40476 20.09577 19.80761 19.53887 19.28824 19.05450 18.83652 18.63322
#> 15 20.40476 20.15177 19.91583 19.69578 19.49057 19.29919 19.12071 18.95426
#> 16 20.40476 19.88497 19.40021 18.94813 18.52651 18.13331 17.76661 17.42462
#> 17 20.40476 19.88168 19.39385 18.93890 18.51461 18.11892 17.74989 17.40574
#> 18 20.40476 21.13990 21.82550 22.46488 23.06118 23.61728 24.13591 24.61958
#> 19 20.40476 21.00486 21.56451 22.08644 22.57320 23.02714 23.45050 23.84532
#> 20 20.40476 20.32634 20.25320 20.18499 20.12138 20.06206 20.00673 19.95513
#> 21 20.40476 20.06284 19.74396 19.44657 19.16922 18.91057 18.66934 18.44438
#>      2.7000   2.5180   2.3483   2.1900   2.0424   1.9048   1.7764   1.6567
#> 1  22.20661 22.36907 22.52057 22.66186 22.79364 22.91653 23.03113 23.13802
#> 2  21.14043 21.20676 21.26862 21.32631 21.38011 21.43028 21.47708 21.52072
#> 3  23.46094 23.73648 23.99345 24.23311 24.45661 24.66505 24.85944 25.04073
#> 4  19.71886 19.65702 19.59935 19.54557 19.49541 19.44863 19.40500 19.36431
#> 5  18.69450 18.54030 18.39650 18.26238 18.13731 18.02067 17.91189 17.81043
#> 6  19.82339 19.77098 19.72209 19.67650 19.63399 19.59434 19.55736 19.52287
#> 7  18.77812 18.63146 18.49469 18.36713 18.24818 18.13724 18.03377 17.93728
#> 8  16.14403 15.75989 15.40163 15.06752 14.75593 14.46534 14.19434 13.94159
#> 9  17.56560 17.30963 17.07090 16.84827 16.64064 16.44700 16.26641 16.09800
#> 10 17.35655 17.08172 16.82542 16.58639 16.36347 16.15558 15.96170 15.78088
#> 11 26.40861 26.94991 27.45473 27.92553 28.36459 28.77407 29.15595 29.51209
#> 12 25.48877 25.94714 26.37462 26.77328 27.14508 27.49182 27.81519 28.11677
#> 13 22.85468 23.07556 23.28156 23.47367 23.65284 23.81993 23.97576 24.12108
#> 14 18.44363 18.26682 18.10192 17.94814 17.80472 17.67096 17.54622 17.42989
#> 15 18.79902 18.65425 18.51924 18.39332 18.27589 18.16638 18.06424 17.96899
#> 16 17.10568 16.80824 16.53084 16.27214 16.03088 15.80587 15.59603 15.40034
#> 17 17.08478 16.78545 16.50630 16.24596 16.00316 15.77673 15.56556 15.36863
#> 18 25.07066 25.49133 25.88365 26.24954 26.59076 26.90898 27.20576 27.48253
#> 19 24.21354 24.55693 24.87718 25.17585 25.45439 25.71416 25.95642 26.18235
#> 20 19.90701 19.86214 19.82028 19.78125 19.74485 19.71090 19.67924 19.64972
#> 21 18.23458 18.03892 17.85644 17.68626 17.52755 17.37954 17.24151 17.11277
#>      1.5450   1.4409   1.3438   1.2532   1.1688   1.0900   1.0165   0.9480
#> 1  23.21294 23.21294 23.21294 23.15933 23.00803 22.86694 22.73536 22.61273
#> 2  21.55130 21.55130 21.55130 21.52224 21.44019 21.36368 21.29233 21.22583
#> 3  25.16780 25.16780 25.16780 25.17624 25.20005 25.22226 25.24298 25.26228
#> 4  19.33579 19.33579 19.33579 19.39360 19.55676 19.70890 19.85080 19.98303
#> 5  17.73932 17.73932 17.73932 17.82853 18.08034 18.31514 18.53412 18.73820
#> 6  19.49870 19.49870 19.49870 19.57400 19.78655 19.98474 20.16958 20.34184
#> 7  17.86965 17.86965 17.86965 17.91563 18.04543 18.16646 18.27934 18.38454
#> 8  13.76444 13.76444 13.76444 13.77772 13.81519 13.85014 13.88273 13.91310
#> 9  15.97995 15.97995 15.97995 15.99115 16.02277 16.05225 16.07975 16.10537
#> 10 15.65414 15.65414 15.65414 15.67954 15.75125 15.81812 15.88048 15.93860
#> 11 29.76172 29.76172 29.76172 29.74952 29.71508 29.68297 29.65302 29.62511
#> 12 28.32816 28.32816 28.32816 28.36615 28.47340 28.57340 28.66667 28.75358
#> 13 24.22295 24.22295 24.22295 24.28034 24.44235 24.59342 24.73430 24.86560
#> 14 17.34835 17.34835 17.34835 17.33065 17.28067 17.23407 17.19061 17.15011
#> 15 17.90223 17.90223 17.90223 17.89630 17.87958 17.86398 17.84944 17.83588
#> 16 15.26317 15.26317 15.26317 15.20626 15.04561 14.89581 14.75611 14.62592
#> 17 15.23058 15.23058 15.23058 15.22688 15.21641 15.20665 15.19755 15.18907
#> 18 27.67654 27.67654 27.67654 27.68468 27.70767 27.72910 27.74909 27.76772
#> 19 26.34071 26.34071 26.34071 26.28281 26.11936 25.96696 25.82482 25.69235
#> 20 19.62902 19.62902 19.62902 19.52582 19.23449 18.96283 18.70948 18.47337
#> 21 17.02254 17.02254 17.02254 16.93263 16.67882 16.44216 16.22144 16.01574
#>      0.8841   0.8245   0.7690   0.7171   0.6688   0.6237   0.5817   0.5425
#> 1  22.49829 22.39155 22.29201 22.19918 22.11261 22.03187 21.95657 21.88635
#> 2  21.16377 21.10589 21.05191 21.00158 20.95463 20.91085 20.87001 20.83193
#> 3  25.28030 25.29710 25.31277 25.32738 25.34101 25.35372 25.36557 25.37663
#> 4  20.10645 20.22155 20.32889 20.42899 20.52235 20.60942 20.69062 20.76634
#> 5  18.92867 19.10630 19.27196 19.42645 19.57053 19.70490 19.83022 19.94708
#> 6  20.50261 20.65254 20.79237 20.92278 21.04440 21.15782 21.26359 21.36224
#> 7  18.48272 18.57428 18.65967 18.73931 18.81358 18.88284 18.94744 19.00768
#> 8  13.94145 13.96789 13.99254 14.01554 14.03698 14.05698 14.07563 14.09303
#> 9  16.12929 16.15159 16.17239 16.19179 16.20988 16.22676 16.24249 16.25717
#> 10 15.99284 16.04342 16.09060 16.13459 16.17562 16.21389 16.24958 16.28286
#> 11 29.59906 29.57477 29.55211 29.53098 29.51128 29.49290 29.47576 29.45978
#> 12 28.83470 28.91036 28.98091 29.04671 29.10808 29.16531 29.21868 29.26845
#> 13 24.98814 25.10243 25.20901 25.30840 25.40110 25.48755 25.56818 25.64337
#> 14 17.11231 17.07705 17.04418 17.01351 16.98492 16.95825 16.93338 16.91019
#> 15 17.82323 17.81144 17.80043 17.79017 17.78060 17.77168 17.76335 17.75559
#> 16 14.50441 14.39109 14.28540 14.18684 14.09492 14.00920 13.92926 13.85470
#> 17 15.18115 15.17377 15.16688 15.16046 15.15448 15.14889 15.14368 15.13882
#> 18 27.78511 27.80133 27.81645 27.83055 27.84371 27.85597 27.86741 27.87808
#> 19 25.56872 25.45342 25.34590 25.24562 25.15210 25.06488 24.98354 24.90768
#> 20 18.25301 18.04751 17.85585 17.67711 17.51041 17.35496 17.20997 17.07476
#> 21 15.82377 15.64473 15.47776 15.32204 15.17682 15.04138 14.91507 14.79728
#>      0.5059   0.4718   0.4400   0.4104   0.3827   0.3569   0.3329   0.3104
#> 1  21.82086 21.77870 21.77864 21.77864 21.77864 21.77864 21.77864 21.77864
#> 2  20.79642 20.77356 20.77353 20.77352 20.77352 20.77352 20.77352 20.77352
#> 3  25.38694 25.39357 25.39358 25.39358 25.39358 25.39358 25.39358 25.39358
#> 4  20.83697 20.88243 20.88249 20.88250 20.88250 20.88250 20.88250 20.88250
#> 5  20.05608 20.12624 20.12633 20.12634 20.12634 20.12634 20.12634 20.12634
#> 6  21.45424 21.51346 21.51354 21.51355 21.51355 21.51355 21.51355 21.51355
#> 7  19.06386 19.10003 19.10007 19.10008 19.10008 19.10008 19.10008 19.10008
#> 8  14.10925 14.11969 14.11970 14.11970 14.11971 14.11971 14.11971 14.11971
#> 9  16.27085 16.27966 16.27967 16.27967 16.27967 16.27967 16.27967 16.27967
#> 10 16.31390 16.33388 16.33390 16.33391 16.33391 16.33391 16.33391 16.33391
#> 11 29.44487 29.43528 29.43526 29.43526 29.43526 29.43526 29.43526 29.43526
#> 12 29.31487 29.34476 29.34479 29.34480 29.34480 29.34480 29.34480 29.34480
#> 13 25.71349 25.75863 25.75869 25.75870 25.75870 25.75870 25.75870 25.75870
#> 14 16.88855 16.87463 16.87461 16.87461 16.87461 16.87461 16.87461 16.87461
#> 15 17.74835 17.74369 17.74368 17.74368 17.74368 17.74368 17.74368 17.74368
#> 16 13.78517 13.74040 13.74035 13.74034 13.74034 13.74034 13.74034 13.74034
#> 17 15.13430 15.13138 15.13138 15.13137 15.13137 15.13137 15.13137 15.13137
#> 18 27.88803 27.89444 27.89445 27.89445 27.89445 27.89445 27.89445 27.89445
#> 19 24.83693 24.79139 24.79133 24.79133 24.79133 24.79132 24.79132 24.79132
#> 20 16.94866 16.86749 16.86738 16.86737 16.86737 16.86737 16.86737 16.86737
#> 21 14.68742 14.61670 14.61661 14.61660 14.61660 14.61660 14.61660 14.61660
#>      0.2895   0.2700   0.2518   0.2348   0.2190   0.2042   0.1905   0.1776
#> 1  21.76395 21.70808 21.65590 21.97673 21.23364 21.23354 21.23343 21.23334
#> 2  20.77500 20.78031 20.78525 21.60050 22.58619 22.58630 22.58641 22.58650
#> 3  25.39664 25.40823 25.41906 24.80864 24.29537 24.29531 24.29526 24.29521
#> 4  20.87481 20.84635 20.81979 20.04002 20.17941 20.17944 20.17946 20.17949
#> 5  20.13735 20.17967 20.21922 20.11422 20.11049 20.11057 20.11065 20.11071
#> 6  21.56492 21.75938 21.94100 22.58767 22.15762 22.15767 22.15772 22.15776
#> 7  19.14572 19.31830 19.47948 19.24959 18.74668 18.74663 18.74658 18.74654
#> 8  14.11495 14.09713 14.08049 14.39271 14.64500 14.64501 14.64501 14.64501
#> 9  16.26576 16.21336 16.16444 16.59892 16.46487 16.46485 16.46483 16.46481
#> 10 16.32695 16.30096 16.27669 17.04768 17.84077 17.84087 17.84097 17.84107
#> 11 29.41272 29.32757 29.24805 28.47450 30.07369 30.07375 30.07380 30.07385
#> 12 29.34793 29.36008 29.37144 29.76481 31.40074 31.40096 31.40117 31.40136
#> 13 25.73249 25.63417 25.54235 25.13861 23.78368 23.78348 23.78330 23.78313
#> 14 16.84387 16.72775 16.61931 16.42471 14.06482 14.06457 14.06433 14.06410
#> 15 17.71562 17.60973 17.51083 17.66049 17.33380 17.33375 17.33370 17.33365
#> 16 13.70158 13.55487 13.41785 12.05833 12.63682 12.63669 12.63657 12.63645
#> 17 15.11425 15.04960 14.98923 14.98417 17.33891 17.33917 17.33941 17.33963
#> 18 27.88796 27.86356 27.84077 27.42124 26.85135 26.85128 26.85121 26.85115
#> 19 24.81575 24.90737 24.99292 26.41869 24.94054 24.94047 24.94039 24.94032
#> 20 16.89869 17.01591 17.12536 16.86479 18.50791 18.50811 18.50830 18.50848
#> 21 14.66308 14.83761 15.00059 14.87299 13.30771 13.30761 13.30751 13.30743
#>      0.1657   0.1545   0.1441   0.1344   0.1253   0.1169   0.1090   0.1017
#> 1  21.23325 21.23317 21.23309 21.23302 21.23295 21.23289 21.28506 21.28504
#> 2  22.58660 22.58668 22.58676 22.58684 22.58691 22.58697 22.73410 22.73417
#> 3  24.29516 24.29512 24.29508 24.29504 24.29501 24.29497 24.06463 24.06446
#> 4  20.17951 20.17953 20.17955 20.17957 20.17959 20.17960 20.38671 20.38677
#> 5  20.11078 20.11084 20.11089 20.11095 20.11100 20.11104 20.15330 20.15333
#> 6  22.15780 22.15784 22.15788 22.15791 22.15794 22.15797 23.23956 23.23973
#> 7  18.74649 18.74645 18.74641 18.74638 18.74635 18.74631 18.24626 18.24611
#> 8  14.64501 14.64501 14.64501 14.64501 14.64501 14.64501 13.99392 13.99374
#> 9  16.46480 16.46478 16.46477 16.46475 16.46474 16.46473 16.04680 16.04674
#> 10 17.84115 17.84123 17.84131 17.84138 17.84145 17.84151 17.55248 17.55252
#> 11 30.07390 30.07395 30.07399 30.07403 30.07407 30.07410 29.61003 29.61001
#> 12 31.40154 31.40171 31.40187 31.40202 31.40216 31.40229 31.16930 31.16940
#> 13 23.78297 23.78281 23.78267 23.78254 23.78242 23.78230 24.35861 24.35868
#> 14 14.06389 14.06369 14.06351 14.06334 14.06318 14.06303 13.79087 13.79074
#> 15 17.33361 17.33357 17.33354 17.33350 17.33347 17.33344 16.87948 16.87942
#> 16 12.63635 12.63625 12.63615 12.63607 12.63598 12.63591 12.58715 12.58704
#> 17 17.33984 17.34004 17.34023 17.34040 17.34056 17.34071 18.06562 18.06590
#> 18 26.85110 26.85104 26.85099 26.85095 26.85090 26.85086 26.37252 26.37237
#> 19 24.94026 24.94020 24.94014 24.94009 24.94004 24.94000 25.73050 25.73058
#> 20 18.50864 18.50880 18.50894 18.50908 18.50920 18.50932 18.12500 18.12497
#> 21 13.30735 13.30727 13.30720 13.30713 13.30707 13.30701 14.10812 14.10828
#>      0.0948   0.0884   0.0825   0.0769   0.0717   0.0669   0.0624   0.0582
#> 1  21.28502 21.28501 21.28499 21.28498 21.28497 21.69198 21.69193 21.69188
#> 2  22.73424 22.73431 22.73437 22.73443 22.73449 23.01631 23.01630 23.01630
#> 3  24.06429 24.06414 24.06399 24.06385 24.06371 23.93562 23.93542 23.93524
#> 4  20.38684 20.38690 20.38696 20.38701 20.38706 20.57094 20.57099 20.57103
#> 5  20.15336 20.15339 20.15342 20.15345 20.15347 20.17085 20.17085 20.17086
#> 6  23.23990 23.24007 23.24022 23.24037 23.24051 23.41460 23.41475 23.41491
#> 7  18.24597 18.24583 18.24570 18.24557 18.24545 18.49307 18.49281 18.49257
#> 8  13.99356 13.99340 13.99324 13.99308 13.99293 13.82467 13.82445 13.82424
#> 9  16.04668 16.04663 16.04658 16.04653 16.04649 15.85490 15.85488 15.85486
#> 10 17.55255 17.55258 17.55261 17.55264 17.55267 17.27822 17.27828 17.27835
#> 11 29.60999 29.60997 29.60995 29.60994 29.60992 29.83729 29.83719 29.83710
#> 12 31.16950 31.16959 31.16968 31.16976 31.16984 30.90572 30.90582 30.90592
#> 13 24.35875 24.35881 24.35887 24.35893 24.35899 23.92822 23.92849 23.92874
#> 14 13.79061 13.79049 13.79037 13.79026 13.79015 14.07870 14.07860 14.07849
#> 15 16.87937 16.87932 16.87927 16.87922 16.87918 16.96631 16.96626 16.96621
#> 16 12.58694 12.58684 12.58674 12.58665 12.58656 12.46353 12.46345 12.46338
#> 17 18.06616 18.06641 18.06666 18.06689 18.06711 18.25571 18.25595 18.25618
#> 18 26.37223 26.37210 26.37197 26.37184 26.37172 26.34555 26.34538 26.34522
#> 19 25.73065 25.73072 25.73079 25.73085 25.73091 25.60872 25.60888 25.60904
#> 20 18.12495 18.12493 18.12491 18.12489 18.12487 17.97334 17.97324 17.97316
#> 21 14.10843 14.10857 14.10871 14.10884 14.10897 13.88579 13.88607 13.88633
#>      0.0542   0.0506   0.0472   0.0440   0.0410   0.0383   0.0357   0.0333
#> 1  21.69182 21.69177 21.69172 21.69166 21.69161 21.69156 21.69150 21.69145
#> 2  23.01630 23.01629 23.01629 23.01630 23.01630 23.01630 23.01631 23.01631
#> 3  23.93507 23.93490 23.93475 23.93460 23.93446 23.93433 23.93421 23.93409
#> 4  20.57107 20.57110 20.57113 20.57116 20.57118 20.57120 20.57122 20.57124
#> 5  20.17087 20.17087 20.17088 20.17089 20.17089 20.17090 20.17091 20.17092
#> 6  23.41505 23.41519 23.41533 23.41546 23.41558 23.41570 23.41581 23.41592
#> 7  18.49233 18.49211 18.49191 18.49171 18.49152 18.49134 18.49117 18.49102
#> 8  13.82405 13.82387 13.82370 13.82354 13.82339 13.82325 13.82312 13.82300
#> 9  15.85484 15.85483 15.85481 15.85480 15.85479 15.85478 15.85477 15.85476
#> 10 17.27842 17.27849 17.27856 17.27862 17.27869 17.27875 17.27881 17.27887
#> 11 29.83701 29.83692 29.83684 29.83676 29.83669 29.83662 29.83655 29.83649
#> 12 30.90602 30.90612 30.90622 30.90631 30.90641 30.90650 30.90659 30.90668
#> 13 23.92898 23.92921 23.92942 23.92962 23.92981 23.92998 23.93015 23.93031
#> 14 14.07839 14.07828 14.07817 14.07807 14.07796 14.07786 14.07775 14.07765
#> 15 16.96617 16.96612 16.96608 16.96603 16.96599 16.96595 16.96590 16.96586
#> 16 12.46330 12.46323 12.46317 12.46310 12.46304 12.46298 12.46293 12.46287
#> 17 18.25640 18.25661 18.25681 18.25700 18.25719 18.25736 18.25753 18.25769
#> 18 26.34508 26.34493 26.34480 26.34467 26.34454 26.34442 26.34431 26.34420
#> 19 25.60918 25.60932 25.60945 25.60958 25.60970 25.60981 25.60992 25.61002
#> 20 17.97308 17.97301 17.97294 17.97288 17.97283 17.97278 17.97273 17.97269
#> 21 13.88658 13.88681 13.88703 13.88724 13.88744 13.88762 13.88779 13.88796
#>      0.0310   0.0290   0.0270   0.0252   0.0235   0.0219   0.0204   0.0190
#> 1  21.69140 21.69135 21.69130 21.69120 21.74204 21.74200 21.74196 21.74192
#> 2  23.01632 23.01632 23.01633 23.01624 23.09651 23.09673 23.09693 23.09713
#> 3  23.93398 23.93387 23.93377 23.93371 23.98521 23.98532 23.98543 23.98552
#> 4  20.57126 20.57127 20.57128 20.57145 20.58327 20.58322 20.58318 20.58314
#> 5  20.17092 20.17093 20.17094 20.17105 20.16213 20.16217 20.16220 20.16224
#> 6  23.41603 23.41613 23.41623 23.41619 23.36110 23.36129 23.36146 23.36162
#> 7  18.49086 18.49072 18.49058 18.49035 18.39418 18.39419 18.39420 18.39420
#> 8  13.82289 13.82278 13.82268 13.82261 13.85256 13.85269 13.85282 13.85294
#> 9  15.85475 15.85475 15.85474 15.85479 15.84390 15.84383 15.84377 15.84370
#> 10 17.27893 17.27899 17.27904 17.27914 17.27352 17.27355 17.27358 17.27360
#> 11 29.83643 29.83637 29.83632 29.83637 29.88936 29.88938 29.88939 29.88941
#> 12 30.90676 30.90684 30.90692 30.90704 30.89757 30.89764 30.89771 30.89778
#> 13 23.93046 23.93060 23.93073 23.93108 23.98179 23.98165 23.98152 23.98139
#> 14 14.07755 14.07745 14.07735 14.07735 14.01137 14.01099 14.01064 14.01030
#> 15 16.96582 16.96578 16.96574 16.96580 16.91040 16.91018 16.90997 16.90978
#> 16 12.46282 12.46277 12.46272 12.46281 12.61313 12.61322 12.61331 12.61338
#> 17 18.25784 18.25799 18.25813 18.25832 18.31631 18.31648 18.31663 18.31678
#> 18 26.34410 26.34400 26.34390 26.34387 26.34245 26.34239 26.34233 26.34228
#> 19 25.61012 25.61021 25.61030 25.61013 25.49957 25.49954 25.49951 25.49948
#> 20 17.97266 17.97262 17.97259 17.97230 17.86813 17.86817 17.86822 17.86826
#> 21 13.88811 13.88826 13.88840 13.88822 13.87548 13.87536 13.87526 13.87516
#>      0.0178   0.0166   0.0155   0.0144   0.0134   0.0125   0.0117   0.0109
#> 1  21.74188 21.74184 21.74181 21.74178 21.74174 21.74171 21.74168 21.74165
#> 2  23.09731 23.09748 23.09764 23.09779 23.09793 23.09806 23.09818 23.09830
#> 3  23.98561 23.98569 23.98577 23.98584 23.98590 23.98596 23.98602 23.98607
#> 4  20.58310 20.58307 20.58303 20.58300 20.58298 20.58295 20.58293 20.58291
#> 5  20.16227 20.16231 20.16234 20.16237 20.16239 20.16242 20.16245 20.16247
#> 6  23.36178 23.36192 23.36206 23.36219 23.36231 23.36242 23.36252 23.36262
#> 7  18.39421 18.39421 18.39421 18.39421 18.39421 18.39421 18.39421 18.39420
#> 8  13.85305 13.85315 13.85324 13.85333 13.85341 13.85349 13.85356 13.85363
#> 9  15.84365 15.84359 15.84354 15.84350 15.84345 15.84341 15.84337 15.84333
#> 10 17.27363 17.27365 17.27367 17.27369 17.27372 17.27374 17.27376 17.27378
#> 11 29.88943 29.88944 29.88946 29.88947 29.88949 29.88950 29.88951 29.88952
#> 12 30.89784 30.89791 30.89797 30.89802 30.89808 30.89813 30.89818 30.89823
#> 13 23.98127 23.98116 23.98105 23.98095 23.98086 23.98077 23.98069 23.98061
#> 14 14.00998 14.00968 14.00940 14.00913 14.00888 14.00864 14.00842 14.00821
#> 15 16.90959 16.90942 16.90926 16.90911 16.90897 16.90883 16.90871 16.90859
#> 16 12.61346 12.61352 12.61358 12.61364 12.61369 12.61374 12.61379 12.61383
#> 17 18.31692 18.31705 18.31718 18.31730 18.31742 18.31753 18.31763 18.31773
#> 18 26.34222 26.34217 26.34213 26.34208 26.34204 26.34200 26.34196 26.34192
#> 19 25.49945 25.49942 25.49939 25.49937 25.49934 25.49932 25.49930 25.49928
#> 20 17.86829 17.86833 17.86837 17.86841 17.86844 17.86847 17.86851 17.86854
#> 21 13.87507 13.87498 13.87490 13.87483 13.87476 13.87469 13.87463 13.87457
#>      0.0102   0.0095   0.0088   0.0082   0.0077   0.0072   0.0067   0.0062
#> 1  21.74162 21.74160 21.74157 21.74155 21.74153 21.74151 21.74148 21.74146
#> 2  23.09841 23.09851 23.09861 23.09870 23.09879 23.09887 23.09894 23.09901
#> 3  23.98612 23.98617 23.98621 23.98625 23.98628 23.98632 23.98635 23.98638
#> 4  20.58289 20.58287 20.58285 20.58284 20.58282 20.58281 20.58280 20.58278
#> 5  20.16249 20.16252 20.16254 20.16256 20.16257 20.16259 20.16261 20.16262
#> 6  23.36272 23.36280 23.36288 23.36296 23.36304 23.36310 23.36317 23.36323
#> 7  18.39420 18.39420 18.39419 18.39419 18.39418 18.39417 18.39417 18.39416
#> 8  13.85369 13.85374 13.85380 13.85384 13.85389 13.85393 13.85397 13.85401
#> 9  15.84330 15.84326 15.84323 15.84320 15.84318 15.84315 15.84313 15.84310
#> 10 17.27379 17.27381 17.27383 17.27385 17.27386 17.27388 17.27389 17.27390
#> 11 29.88954 29.88955 29.88956 29.88956 29.88957 29.88958 29.88959 29.88959
#> 12 30.89828 30.89832 30.89837 30.89841 30.89845 30.89848 30.89852 30.89855
#> 13 23.98054 23.98047 23.98041 23.98035 23.98030 23.98025 23.98020 23.98016
#> 14 14.00801 14.00783 14.00765 14.00749 14.00733 14.00718 14.00704 14.00691
#> 15 16.90848 16.90837 16.90828 16.90818 16.90810 16.90801 16.90794 16.90786
#> 16 12.61387 12.61390 12.61394 12.61397 12.61400 12.61402 12.61405 12.61407
#> 17 18.31782 18.31791 18.31800 18.31808 18.31816 18.31823 18.31830 18.31837
#> 18 26.34189 26.34185 26.34182 26.34179 26.34176 26.34173 26.34171 26.34168
#> 19 25.49926 25.49924 25.49922 25.49920 25.49919 25.49917 25.49916 25.49915
#> 20 17.86857 17.86859 17.86862 17.86864 17.86867 17.86869 17.86871 17.86873
#> 21 13.87452 13.87448 13.87443 13.87439 13.87435 13.87432 13.87429 13.87426
#>      0.0058   0.0054   0.0051   0.0047
#> 1  21.74145 21.74143 21.74141 21.74139
#> 2  23.09908 23.09914 23.09920 23.09926
#> 3  23.98640 23.98643 23.98645 23.98647
#> 4  20.58277 20.58276 20.58276 20.58275
#> 5  20.16264 20.16265 20.16267 20.16268
#> 6  23.36329 23.36334 23.36339 23.36344
#> 7  18.39415 18.39414 18.39414 18.39413
#> 8  13.85404 13.85407 13.85410 13.85412
#> 9  15.84308 15.84306 15.84304 15.84303
#> 10 17.27392 17.27393 17.27394 17.27396
#> 11 29.88960 29.88961 29.88961 29.88962
#> 12 30.89859 30.89862 30.89865 30.89868
#> 13 23.98012 23.98008 23.98004 23.98001
#> 14 14.00679 14.00667 14.00656 14.00646
#> 15 16.90780 16.90773 16.90767 16.90762
#> 16 12.61409 12.61411 12.61413 12.61415
#> 17 18.31843 18.31849 18.31855 18.31860
#> 18 26.34166 26.34164 26.34161 26.34159
#> 19 25.49913 25.49912 25.49911 25.49910
#> 20 17.86875 17.86877 17.86878 17.86880
#> 21 13.87423 13.87420 13.87418 13.87416
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 22.8 21.4 18.1 24.4 17.8 16.4 17.3 15.2 30.4 33.9 21.5 15.5 15.2
#> [16] 13.3 19.2 27.3 26.0 15.8 15.0
#> 
#> $X
#>               am       carb        cyl        disp       drat       gear
#>  [1,]  1.1547005  0.8385567 -0.1097643 -0.56212745  0.4803537  0.4677072
#>  [2,]  1.1547005  0.8385567 -0.1097643 -0.56212745  0.4803537  0.4677072
#>  [3,]  1.1547005 -0.9831354 -1.2622890 -1.08105819  0.3926369  0.4677072
#>  [4,] -0.8660254 -0.9831354 -0.1097643  0.41585741 -0.9582012 -0.9354143
#>  [5,] -0.8660254 -0.9831354 -0.1097643  0.08653598 -1.5195884 -0.9354143
#>  [6,] -0.8660254 -0.3759047 -1.2622890 -0.69485397  0.1119433  0.4677072
#>  [7,] -0.8660254  0.8385567 -0.1097643 -0.48628373  0.5154404  0.4677072
#>  [8,] -0.8660254  0.2313260  1.0427605  0.59349139 -0.9757445 -0.9354143
#>  [9,] -0.8660254  0.2313260  1.0427605  0.59349139 -0.9757445 -0.9354143
#> [10,] -0.8660254  0.2313260  1.0427605  0.59349139 -0.9757445 -0.9354143
#> [11,]  1.1547005 -0.3759047 -1.2622890 -1.40339402  2.2873189  0.4677072
#> [12,]  1.1547005 -0.9831354 -1.2622890 -1.44929943  1.0417409  0.4677072
#> [13,] -0.8660254 -0.9831354 -1.2622890 -0.96030700  0.1294866 -0.9354143
#> [14,] -0.8660254 -0.3759047  1.0427605  1.01462365 -1.5195884 -0.9354143
#> [15,] -0.8660254 -0.3759047  1.0427605  0.87491153 -0.8353977 -0.9354143
#> [16,] -0.8660254  0.8385567  1.0427605  1.33396565  0.1821167 -0.9354143
#> [17,] -0.8660254 -0.3759047  1.0427605  1.83293751 -0.9582012 -0.9354143
#> [18,]  1.1547005 -0.9831354 -1.2622890 -1.37046187  0.7961340  0.4677072
#> [19,]  1.1547005 -0.3759047 -1.2622890 -0.95831111  1.4101513  1.8708287
#> [20,]  1.1547005  0.8385567  1.0427605  1.34394508  1.0417409  1.8708287
#> [21,]  1.1547005  3.2674794  1.0427605  0.84497322 -0.1512070  1.8708287
#>               hp       qsec         vs         wt
#>  [1,] -0.4237643 -0.7907755 -0.8660254 -0.5951709
#>  [2,] -0.4237643 -0.4556576 -0.8660254 -0.2430007
#>  [3,] -0.6627593  0.4958376  1.1547005 -1.0094888
#>  [4,] -0.4237643  0.9925301  1.1547005  0.2265595
#>  [5,] -0.4940569  1.4593013  1.1547005  0.5649191
#>  [6,] -1.0985737  1.3276479  1.1547005  0.1920330
#>  [7,] -0.2410034  0.6693807  1.1547005  0.5372979
#>  [8,]  0.5603329 -0.2282563 -0.8660254  1.4073655
#>  [9,]  0.5603329 -0.1085713 -0.8660254  0.9378052
#> [10,]  0.5603329  0.1307985 -0.8660254  1.0068582
#> [11,] -1.2391590  0.4419794  1.1547005 -1.9831358
#> [12,] -1.0563981  1.2678054  1.1547005 -1.6793027
#> [13,] -0.6065252  1.3336321  1.1547005 -0.8092352
#> [14,]  0.1385769 -0.5454213 -0.8660254  0.6477827
#> [15,]  0.1385769 -0.2880987 -0.8660254  0.5303926
#> [16,]  1.4741373 -1.4191214 -0.8660254  1.0897218
#> [17,]  0.4900402 -0.4377049 -0.8660254  1.0966271
#> [18,] -1.0423396  0.6693807  1.1547005 -1.5411967
#> [19,] -0.6908763 -0.6471535 -0.8660254 -1.2580795
#> [20,]  1.7412494 -1.9636878 -0.8660254  0.1644119
#> [21,]  2.7394051 -1.9038454 -0.8660254  0.7168357
#> attr(,"center")
#>  [1]   0.4285714   2.6190476   6.1904762 216.3285714   3.6261905   3.6666667
#>  [7] 140.1428571  17.7814286   0.4285714   3.0509524
#> attr(,"scale")
#>  [1]   0.4948717   1.6468206   1.7353207 100.2060503   0.5700165   0.7126966
#>  [7]  71.1311905   1.6710541   0.4948717   0.7240817
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 36
#> 
#> $lambda.min
#> [1] 0.4103706
#> 
#> $null.dev
#> [1] 28.75379
#> 
#> $Bias
#> [1] 2.078832
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 11.62254