Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 48.772198 45.192299 39.644413 33.884821 29.107629 25.169227 21.945767
#>   [8] 19.330515 17.231558 15.569809 14.277281 13.295586 12.574634 12.071508
#>  [15] 11.749485 11.576018 11.277639 11.096535 11.095352 11.095352 11.126800
#>  [22] 11.143693 11.171377 11.152391 11.078764 10.923367 10.751908 10.384591
#>  [29] 10.138649  9.962218  9.843651  9.773059 10.868198 10.986836 10.877914
#>  [36] 10.716986 10.422991 10.372075 15.480360 15.421797 15.370885 14.582644
#>  [43] 14.128288 14.196547 15.090084 14.818928 14.422474 14.406291 14.289105
#>  [50] 14.882182 14.893626 13.813318 14.726402 14.718392 14.953833 14.707013
#>  [57] 14.729566 14.729945 14.730290 14.893108 14.913872 14.899457 14.899341
#>  [64] 14.875071 14.874853 14.874656 15.622338 15.572983 16.186910 16.186694
#>  [71] 16.138638 16.042541 16.042512 16.042485 16.042460 16.006222 16.006595
#>  [78] 16.006940 16.763914 16.765003 16.766026 16.370870 16.371773 16.362200
#>  [85] 16.916417 16.917251 16.918070 16.918876 16.919398 16.994048 16.846143
#>  [92] 16.846858 16.847551 16.848224 16.848877 16.849512 16.850130 16.850731
#>  [99] 16.851816 16.957914
#> 
#> $cvse
#>   [1] 11.948850 11.451716 10.268977  8.942887  7.846415  6.936707  6.178305
#>   [8]  5.542794  5.008306  4.558858  4.183515  3.875403  3.630581  3.446789
#>  [15]  3.322172  3.254029  3.123345  3.055303  3.055514  3.055514  3.059153
#>  [22]  3.071628  3.077125  3.046402  3.001788  2.919521  2.854610  2.788873
#>  [29]  2.744425  2.716879  2.698876  2.685923  2.671978  2.638279  2.303838
#>  [36]  2.187958  2.091841  2.065222  6.390820  6.391505  6.391914  6.430360
#>  [43]  6.421035  6.417561  6.463385  6.430361  6.419166  6.427582  6.426826
#>  [50]  6.473965  6.470077  5.264033  5.355167  5.352678  4.963660  4.954283
#>  [57]  4.872617  4.872716  4.872811  4.866430  4.863564  4.859515  4.859572
#>  [64]  4.867361  4.867418  4.867471  4.969081  4.956297  4.917569  4.917629
#>  [71]  4.917817  4.923017  4.923080  4.923139  4.923196  4.918183  4.918271
#>  [78]  4.918353  5.336006  5.336482  5.336945  5.321381  5.321830  5.300467
#>  [85]  5.744946  5.745499  5.746043  5.746577  5.747092  5.759854  5.765726
#>  [92]  5.766264  5.766790  5.767306  5.767811  5.768306  5.768790  5.769266
#>  [99]  5.769823  5.788852
#> 
#> $fold
#>  [1]  5 10  3  4  2  6  9  1  7 10  3  8  1  6  4  8  1  9  2  5  7
#> 
#> $lambda
#>   [1] 6.012781129 5.607539805 5.229610390 4.877152152 4.548448421 4.241898220
#>   [7] 3.956008477 3.689386746 3.440734427 3.208840443 2.992575336 2.790885775
#>  [13] 2.602789414 2.427370119 2.263773497 2.111202740 1.968914741 1.836216477
#>  [19] 1.712461632 1.597047450 1.489411797 1.389030427 1.295414425 1.208107828
#>  [25] 1.126685404 1.050750579 0.979933506 0.913889268 0.852296190 0.794854279
#>  [31] 0.741283761 0.691323716 0.644730812 0.601278113 0.560753980 0.522961039
#>  [37] 0.487715215 0.454844842 0.424189822 0.395600848 0.368938675 0.344073443
#>  [43] 0.320884044 0.299257534 0.279088577 0.260278941 0.242737012 0.226377350
#>  [49] 0.211120274 0.196891474 0.183621648 0.171246163 0.159704745 0.148941180
#>  [55] 0.138903043 0.129541442 0.120810783 0.112668540 0.105075058 0.097993351
#>  [61] 0.091388927 0.085229619 0.079485428 0.074128376 0.069132372 0.064473081
#>  [67] 0.060127811 0.056075398 0.052296104 0.048771522 0.045484484 0.042418982
#>  [73] 0.039560085 0.036893867 0.034407344 0.032088404 0.029925753 0.027908858
#>  [79] 0.026027894 0.024273701 0.022637735 0.021112027 0.019689147 0.018362165
#>  [85] 0.017124616 0.015970474 0.014894118 0.013890304 0.012954144 0.012081078
#>  [91] 0.011266854 0.010507506 0.009799335 0.009138893 0.008522962 0.007948543
#>  [97] 0.007412838 0.006913237 0.006447308 0.006012781
#> 
#> $fit
#> $beta
#>               6.0128     5.6075   5.2296    4.8772    4.5484    4.2419
#> (Intercept) 21.28571 22.9938029 24.58677 26.072380 27.457864 28.749970
#> am           0.00000  0.0000000  0.00000  0.000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.00000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.00000  0.000000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.00000  0.000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.00000  0.000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.00000  0.000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.00000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.00000  0.000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.00000  0.000000  0.000000  0.000000
#> wt           0.00000 -0.5382067 -1.04014 -1.508245 -1.944801 -2.351934
#>                3.9560    3.6894    3.4407    3.2088    2.9926    2.7909
#> (Intercept) 29.954993 31.078801 32.126868 33.104300 34.015855 34.865975
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.731629 -3.085733 -3.415971 -3.723953 -4.011178 -4.279044
#>                2.6028    2.4274   2.2638    2.1112    1.9689    1.8362
#> (Intercept) 35.658799 36.398190 37.08775 37.730833 38.181594 38.181594
#> am           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> wt          -4.528858 -4.761835 -4.97911 -5.181741 -5.323773 -5.323773
#>                1.7125    1.5970    1.4894    1.3890    1.2954     1.2081
#> (Intercept) 38.181594 38.181594 38.181594 38.181594 38.181594 36.9695810
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.0632371
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> wt          -5.323773 -5.323773 -5.323773 -5.323773 -5.323773 -5.3044452
#>                 1.1267     1.0508     0.9799     0.9139     0.8523     0.7949
#> (Intercept) 38.2551311 38.2959287 38.3339728 38.3694528 38.4025416 38.4334004
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb        -0.1732228 -0.2693246 -0.3589406 -0.4425167 -0.5204601 -0.5931503
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.1935962 -5.1213760 -5.0540299 -4.9912228 -4.9326486 -4.8780221
#>                 0.7413     0.6913     0.6447     0.6013    0.5608     0.5230
#> (Intercept) 38.4621794 38.4890187 38.5140492 38.5373927 34.418368 30.6063970
#> am           0.0000000  0.0000000  0.0000000  0.0000000  3.067794  3.5732567
#> carb        -0.6609415 -0.7241638 -0.7831251 -0.8381126 -1.257245 -1.1902808
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.1604681
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> wt          -4.8270772 -4.7795659 -4.7352567 -4.6939338 -3.439291 -3.2857464
#>                 0.4877     0.4548     0.4242     0.3956     0.3689     0.3441
#> (Intercept) 28.1250776 25.8116878 23.6542295 21.6479163 20.9038208 20.9023696
#> am           3.7289170  3.8737694  4.0088573  4.1344410  4.1814556  4.1815042
#> carb        -1.1175020 -1.0495771 -0.9862305 -0.9273867 -0.9054679 -0.9053522
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.2769439  0.3855553  0.4868459  0.5810448  0.6159468  0.6160164
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -3.2571617 -3.2306436 -3.2059131 -3.1828773 -3.1742814 -3.1743322
#>                 0.3209     0.2993     0.2791     0.2603     0.2427    0.2264
#> (Intercept) 20.9017151 20.9014412 20.9013434 20.9013229 19.6765827 12.582855
#> am           4.1814928  4.1814617  4.1814282  4.1813986  4.0907105  2.973752
#> carb        -0.9052888 -0.9052535 -0.9052333 -0.9052215 -0.9340360 -1.491840
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.3906297  1.332826
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  1.657514
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> qsec         0.6160500  0.6160658  0.6160729  0.6160759  0.5889988  0.380316
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> wt          -3.1743731 -3.1744043 -3.1744276 -3.1744446 -3.0515945 -2.064905
#>                 0.2111     0.1969      0.1836      0.1712     0.1597
#> (Intercept) 12.5817308 12.5807642 -14.9080275 -14.9137707 -14.919012
#> am           2.9732742  2.9728663   3.3089419   3.3089884   3.309031
#> carb        -1.4921196 -1.4923584  -2.0119357  -2.0121356  -2.012318
#> cyl          0.0000000  0.0000000   1.6086259   1.6089678   1.609280
#> disp         0.0000000  0.0000000   0.0000000   0.0000000   0.000000
#> drat         1.3330111  1.3331701   2.4261849   2.4263468   2.426495
#> gear         1.6581510  1.6586961   3.2080925   3.2085959   3.209055
#> hp           0.0000000  0.0000000   0.0000000   0.0000000   0.000000
#> qsec         0.3802015  0.3801037   0.9856791   0.9857784   0.985869
#> vs           0.0000000  0.0000000   0.0000000   0.0000000   0.000000
#> wt          -2.0645647 -2.0642730  -2.4925257  -2.4925233  -2.492521
#>                  0.1489     0.1389      0.1295      0.1208      0.1127
#> (Intercept) -14.9237945 -14.928159 -14.9321420 -14.9357768 -14.9390938
#> am            3.3090692   3.309104   3.3091362   3.3091653   3.3091917
#> carb         -2.0124846  -2.012637  -2.0127754  -2.0129020  -2.0130176
#> cyl           1.6095644   1.609824   1.6100613   1.6102776   1.6104750
#> disp          0.0000000   0.000000   0.0000000   0.0000000   0.0000000
#> drat          2.4266294   2.426753   2.4268649   2.4269674   2.4270610
#> gear          3.2094747   3.209857   3.2102068   3.2105257   3.2108167
#> hp            0.0000000   0.000000   0.0000000   0.0000000   0.0000000
#> qsec          0.9859516   0.986027   0.9860958   0.9861586   0.9862158
#> vs            0.0000000   0.000000   0.0000000   0.0000000   0.0000000
#> wt           -2.4925191  -2.492517  -2.4925154  -2.4925138  -2.4925123
#>                  0.1051      0.0980      0.0914      0.0852      0.0795
#> (Intercept) -14.9421209 -15.5249588 -15.5284255 -15.5316612 -15.5346792
#> am            3.3092159   3.3953384   3.3952100   3.3950964   3.3949959
#> carb         -2.0131231  -2.1406731  -2.1409919  -2.1412839  -2.1415515
#> cyl           1.6106551   1.7586497   1.7589110   1.7591541   1.7593803
#> disp          0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> drat          2.4271465   2.5208152   2.5210586   2.5212815   2.5214857
#> gear          3.2110824   3.4387848   3.4394569   3.4400735   3.4406393
#> hp            0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> qsec          0.9862681   0.8437254   0.8436186   0.8435231   0.8434378
#> vs            0.0000000   1.0044528   1.0050595   1.0056136   1.0061197
#> wt           -2.4925108  -2.2203153  -2.2199815  -2.2196787  -2.2194040
#>                   0.0741       0.0691       0.0645       0.0601      0.0561
#> (Intercept) -12.49464676 -12.49662204 -12.49835286 -12.49986168 -12.5011696
#> am            3.64279155   3.64320296   3.64355945   3.64386734   3.6441323
#> carb         -1.85700505  -1.85678296  -1.85658910  -1.85642012  -1.8562730
#> cyl           1.88264567   1.88272201   1.88278039   1.88282343   1.8828534
#> disp          0.00000000   0.00000000   0.00000000   0.00000000   0.0000000
#> drat          2.23793939   2.23775245   2.23759187   2.23745429   2.2373368
#> gear          3.16934585   3.16889567   3.16849588   3.16814121   3.1678269
#> hp           -0.01197606  -0.01197467  -0.01197323  -0.01197177  -0.0119703
#> qsec          0.80932217   0.80958937   0.80982600   0.81003507   0.8102194
#> vs            0.77087007   0.77032600   0.76983724   0.76939879   0.7690060
#> wt           -2.26488881  -2.26539208  -2.26583612  -2.26622710  -2.2665706
#>                   0.0523       0.0488      0.0455       0.0424       0.0396
#> (Intercept) -12.50229654 -12.50326083 -12.5040796 -10.36230580 -10.35888762
#> am            3.64435958   3.64455365   3.6447187   3.56425544   3.56424579
#> carb         -1.85614521  -1.85603427  -1.8559381  -1.39231835  -1.39206010
#> cyl           1.88287245   1.88288227   1.8828845   1.38897469   1.38867237
#> disp          0.00000000   0.00000000   0.0000000   0.01600920   0.01601394
#> drat          2.23723663   2.23715159   2.2370796   2.16965141   2.16946325
#> gear          3.16754854   3.16730231   3.1670846   3.05961237   3.05929996
#> hp           -0.01196886  -0.01196746  -0.0119661  -0.01935859  -0.01936100
#> qsec          0.81038142   0.81052357   0.8106479   0.84531244   0.84531632
#> vs            0.76865460   0.76834063   0.7680605   1.43803210   1.43804554
#> wt           -2.26687175  -2.26713516  -2.2673650  -3.32978124  -3.33018836
#>                   0.0369       0.0344       0.0321       0.0299       0.0279
#> (Intercept) -10.35555300 -10.35229990 -10.34912632 -10.34603031 -10.34300996
#> am            3.56423637   3.56422715   3.56421815   3.56420936   3.56420076
#> carb         -1.39180818  -1.39156244  -1.39132271  -1.39108887  -1.39086075
#> cyl           1.38837743   1.38808969   1.38780899   1.38753515   1.38726800
#> disp          0.01601857   0.01602308   0.01602748   0.01603178   0.01603597
#> drat          2.16927970   2.16910065   2.16892598   2.16875559   2.16858937
#> gear          3.05899521   3.05869792   3.05840792   3.05812503   3.05784907
#> hp           -0.01936335  -0.01936565  -0.01936788  -0.01937007  -0.01937219
#> qsec          0.84532011   0.84532379   0.84532738   0.84533087   0.84533427
#> vs            1.43805867   1.43807149   1.43808401   1.43809623   1.43810816
#> wt           -3.33058551  -3.33097293  -3.33135086  -3.33171952  -3.33207916
#>                   0.0260       0.0243       0.0226       0.0211       0.0197
#> (Intercept) -10.34006343 -10.33718891 -10.33438463 -10.33164888 -10.32897999
#> am            3.56419237   3.56418417   3.56417616   3.56416834   3.56416070
#> carb         -1.39063822  -1.39042114  -1.39020938  -1.39000281  -1.38980130
#> cyl           1.38700738   1.38675313   1.38650509   1.38626311   1.38602705
#> disp          0.01604005   0.01604404   0.01604793   0.01605172   0.01605542
#> drat          2.16842723   2.16826905   2.16811474   2.16796421   2.16781737
#> gear          3.05757986   3.05731725   3.05706106   3.05681115   3.05656736
#> hp           -0.01937427  -0.01937630  -0.01937827  -0.01938020  -0.01938208
#> qsec          0.84533758   0.84534081   0.84534395   0.84534701   0.84534998
#> vs            1.43811981   1.43813117   1.43814227   1.43815309   1.43816365
#> wt           -3.33242998  -3.33277221  -3.33310605  -3.33343171  -3.33374940
#>                   0.0184       0.0171       0.0160       0.0149       0.0139
#> (Intercept) -10.32637631 -10.32383627 -10.32135829 -10.31894087 -10.31658253
#> am            3.56415323   3.56414595   3.56413883   3.56413188   3.56412510
#> carb         -1.38960472  -1.38941295  -1.38922589  -1.38904340  -1.38886538
#> cyl           1.38579676   1.38557209   1.38535292   1.38513911   1.38493052
#> disp          0.01605903   0.01606256   0.01606599   0.01606934   0.01607261
#> drat          2.16767412   2.16753438   2.16739805   2.16726506   2.16713533
#> gear          3.05632953   3.05609753   3.05587120   3.05565041   3.05543502
#> hp           -0.01938391  -0.01938570  -0.01938745  -0.01938915  -0.01939081
#> qsec          0.84535289   0.84535571   0.84535847   0.84536115   0.84536376
#> vs            1.43817396   1.43818401   1.43819382   1.43820339   1.43821272
#> wt           -3.33405930  -3.33436161  -3.33465651  -3.33494419  -3.33522482
#>                   0.0130       0.0121       0.0113       0.0105       0.0098
#> (Intercept) -10.31428183 -10.31203735 -10.30984773 -10.30771163 -10.30562775
#> am            3.56411848   3.56411201   3.56410570   3.56409954   3.56409353
#> carb         -1.38869172  -1.38852231  -1.38835705  -1.38819584  -1.38803857
#> cyl           1.38472703   1.38452852   1.38433487   1.38414595   1.38396165
#> disp          0.01607580   0.01607891   0.01608195   0.01608491   0.01608780
#> drat          2.16700877   2.16688531   2.16676486   2.16664737   2.16653275
#> gear          3.05522490   3.05501993   3.05481997   3.05462490   3.05443460
#> hp           -0.01939243  -0.01939401  -0.01939555  -0.01939705  -0.01939852
#> qsec          0.84536631   0.84536879   0.84537121   0.84537357   0.84537586
#> vs            1.43822182   1.43823070   1.43823936   1.43824780   1.43825604
#> wt           -3.33549858  -3.33576563  -3.33602613  -3.33628026  -3.33652816
#>                   0.0091       0.0085       0.0079       0.0074       0.0069
#> (Intercept) -10.30359480 -10.30161155 -10.29967678 -10.29778932 -10.29594800
#> am            3.56408766   3.56408193   3.56407634   3.56407088   3.56406556
#> carb         -1.38788515  -1.38773549  -1.38758949  -1.38744707  -1.38730813
#> cyl           1.38378185   1.38360646   1.38343535   1.38326843   1.38310559
#> disp          0.01609061   0.01609336   0.01609604   0.01609866   0.01610121
#> drat          2.16642093   2.16631186   2.16620544   2.16610164   2.16600037
#> gear          3.05424896   3.05406785   3.05389118   3.05371883   3.05355070
#> hp           -0.01939995  -0.01940135  -0.01940271  -0.01940404  -0.01940533
#> qsec          0.84537810   0.84538028   0.84538241   0.84538448   0.84538650
#> vs            1.43826407   1.43827190   1.43827953   1.43828698   1.43829424
#> wt           -3.33676999  -3.33700589  -3.33723601  -3.33746050  -3.33767949
#>                  0.0064       0.0060
#> (Intercept) -10.2941517 -10.29239933
#> am            3.5640604   3.56405529
#> carb         -1.3871726  -1.38704038
#> cyl           1.3829467   1.38279177
#> disp          0.0161037   0.01610613
#> drat          2.1659016   2.16580521
#> gear          3.0533867   3.05322667
#> hp           -0.0194066  -0.01940783
#> qsec          0.8453885   0.84539039
#> vs            1.4383013   1.43830822
#> wt           -3.3378931  -3.33810150
#> 
#> $iter
#>   [1]   0   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1   8  17   6   6   6   6   6   6   6   6   6  90  19
#>  [37]  17  17  17  16   8   1   1   1   1   1  48 104   1   1  71   1   1   1
#>  [55]   1   1   1   1   1  56   1   1   1  60   1   1   1   1   1   1   1 186
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 6.012781129 5.607539805 5.229610390 4.877152152 4.548448421 4.241898220
#>   [7] 3.956008477 3.689386746 3.440734427 3.208840443 2.992575336 2.790885775
#>  [13] 2.602789414 2.427370119 2.263773497 2.111202740 1.968914741 1.836216477
#>  [19] 1.712461632 1.597047450 1.489411797 1.389030427 1.295414425 1.208107828
#>  [25] 1.126685404 1.050750579 0.979933506 0.913889268 0.852296190 0.794854279
#>  [31] 0.741283761 0.691323716 0.644730812 0.601278113 0.560753980 0.522961039
#>  [37] 0.487715215 0.454844842 0.424189822 0.395600848 0.368938675 0.344073443
#>  [43] 0.320884044 0.299257534 0.279088577 0.260278941 0.242737012 0.226377350
#>  [49] 0.211120274 0.196891474 0.183621648 0.171246163 0.159704745 0.148941180
#>  [55] 0.138903043 0.129541442 0.120810783 0.112668540 0.105075058 0.097993351
#>  [61] 0.091388927 0.085229619 0.079485428 0.074128376 0.069132372 0.064473081
#>  [67] 0.060127811 0.056075398 0.052296104 0.048771522 0.045484484 0.042418982
#>  [73] 0.039560085 0.036893867 0.034407344 0.032088404 0.029925753 0.027908858
#>  [79] 0.026027894 0.024273701 0.022637735 0.021112027 0.019689147 0.018362165
#>  [85] 0.017124616 0.015970474 0.014894118 0.013890304 0.012954144 0.012081078
#>  [91] 0.011266854 0.010507506 0.009799335 0.009138893 0.008522962 0.007948543
#>  [97] 0.007412838 0.006913237 0.006447308 0.006012781
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 24
#> 
#> $loss
#>   [1] 943.08571 797.33761 675.39765 573.83965 489.70562 420.44328 363.85184
#>   [8] 318.03494 281.35982 252.42181 230.01365 213.09877 200.78814 192.32022
#>  [15] 187.04359 184.40182 183.86144 183.86144 183.86144 183.86144 183.86144
#>  [22] 183.86144 183.86144 177.75249 168.97373 162.06295 156.48586 152.03892
#>  [29] 148.54773 145.86241 143.85434 142.41324 141.44466 140.86787 118.56377
#>  [36] 112.10616 108.72038 106.45738 105.12456 104.55861 104.51372 104.51372
#>  [43] 104.51372 104.51372 104.51372 104.51372 102.67011  94.78253  94.78252
#>  [50]  94.78250  83.33337  83.33335  83.33334  83.33333  83.33332  83.33332
#>  [57]  83.33331  83.33331  83.33330  82.28287  82.28285  82.28283  82.28282
#>  [64]  80.91000  80.90998  80.90997  80.90995  80.90994  80.90994  80.90993
#>  [71]  80.90993  78.86078  78.86076  78.86073  78.86071  78.86069  78.86067
#>  [78]  78.86065  78.86063  78.86061  78.86059  78.86057  78.86056  78.86054
#>  [85]  78.86053  78.86052  78.86050  78.86049  78.86048  78.86047  78.86046
#>  [92]  78.86045  78.86044  78.86043  78.86042  78.86041  78.86040  78.86040
#>  [99]  78.86039  78.86038
#> 
#> $linear.predictors
#>      6.0128   5.6075   5.2296   4.8772   4.5484   4.2419   3.9560   3.6894
#> 1  21.28571 21.14237 21.00869 20.88402 20.76775 20.65932 20.55819 20.46388
#> 2  21.28571 21.13161 20.98789 20.85385 20.72885 20.61228 20.50356 20.40217
#> 3  21.28571 21.27692 21.26873 21.26108 21.25395 21.24730 21.24110 21.23531
#> 4  21.28571 21.29845 21.31033 21.32141 21.33174 21.34138 21.35036 21.35874
#> 5  21.28571 21.14237 21.00869 20.88402 20.76775 20.65932 20.55819 20.46388
#> 6  21.28571 21.14237 21.00869 20.88402 20.76775 20.65932 20.55819 20.46388
#> 7  21.28571 20.16822 19.12604 18.15410 17.24766 16.40231 15.61394 14.87870
#> 8  21.28571 20.07457 18.94505 17.89166 16.90926 15.99308 15.13864 14.34179
#> 9  21.28571 20.11709 19.02722 18.01081 17.06290 16.17888 15.35444 14.58556
#> 10 21.28571 21.80975 22.29846 22.75424 23.17930 23.57571 23.94541 24.29019
#> 11 21.28571 22.12460 22.90695 23.63657 24.31701 24.95160 25.54341 26.09534
#> 12 21.28571 22.00619 22.67812 23.30475 23.88915 24.43417 24.94245 25.41648
#> 13 21.28571 21.66712 22.02283 22.35456 22.66393 22.95245 23.22153 23.47247
#> 14 21.28571 20.92709 20.59263 20.28072 19.98983 19.71854 19.46554 19.22959
#> 15 21.28571 20.92440 20.58743 20.27318 19.98010 19.70678 19.45188 19.21416
#> 16 21.28571 21.95237 22.57410 23.15393 23.69467 24.19898 24.66929 25.10791
#> 17 21.28571 21.84204 22.36087 22.84474 23.29599 23.71683 24.10931 24.47533
#> 18 21.28571 22.17950 23.01304 23.79041 24.51538 25.19149 25.82204 26.41009
#> 19 21.28571 21.50297 21.70558 21.89454 22.07077 22.23511 22.38838 22.53132
#> 20 21.28571 21.07241 20.87347 20.68795 20.51492 20.35356 20.20308 20.06274
#> 21 21.28571 21.49759 21.69518 21.87946 22.05132 22.21159 22.36107 22.50046
#>      3.4407   3.2088   2.9926   2.7909   2.6028   2.4274   2.2638   2.1112
#> 1  20.37593 20.29390 20.21740 20.14606 20.07953 20.01748 19.95961 19.90564
#> 2  20.30761 20.21942 20.13718 20.06048 19.98895 19.92224 19.86003 19.80201
#> 3  21.22992 21.22489 21.22020 21.21582 21.21174 21.20794 21.20439 21.20108
#> 4  21.36656 21.37385 21.38065 21.38699 21.39290 21.39841 21.40355 21.40835
#> 5  20.37593 20.29390 20.21740 20.14606 20.07953 20.01748 19.95961 19.90564
#> 6  20.37593 20.29390 20.21740 20.14606 20.07953 20.01748 19.95961 19.90564
#> 7  14.19302 13.55355 12.95717 12.40099 11.88230 11.39856 10.94742 10.52669
#> 8  13.59864 12.90558 12.25923 11.65644 11.09427 10.57000 10.08106  9.62507
#> 9  13.86850 13.19977 12.57611 11.99448 11.45205 10.94618 10.47441 10.03443
#> 10 24.61173 24.91160 25.19126 25.45208 25.69531 25.92215 26.13371 26.33100
#> 11 26.61007 27.09012 27.53780 27.95532 28.34469 28.70783 29.04649 29.36232
#> 12 25.85856 26.27085 26.65534 27.01393 27.34835 27.66022 27.95108 28.22234
#> 13 23.70650 23.92476 24.12830 24.31813 24.49516 24.66027 24.81424 24.95784
#> 14 19.00954 18.80432 18.61293 18.43444 18.26799 18.11275 17.96797 17.83295
#> 15 18.99246 18.78570 18.59288 18.41305 18.24534 18.08894 17.94307 17.80704
#> 16 25.51696 25.89845 26.25423 26.58602 26.89546 27.18404 27.45317 27.70416
#> 17 24.81669 25.13504 25.43193 25.70882 25.96704 26.20786 26.43245 26.64191
#> 18 26.95850 27.46996 27.94694 28.39178 28.80664 29.19353 29.55436 29.89086
#> 19 22.66463 22.78895 22.90489 23.01302 23.11386 23.20791 23.29561 23.37741
#> 20 19.93185 19.80979 19.69595 19.58979 19.49078 19.39844 19.31233 19.23202
#> 21 22.63047 22.75171 22.86478 22.97023 23.06857 23.16029 23.24582 23.32559
#>       1.9689    1.8362    1.7125    1.5970    1.4894    1.3890    1.2954
#> 1  19.867816 19.867816 19.867816 19.867816 19.867816 19.867816 19.867816
#> 2  19.761341 19.761341 19.761341 19.761341 19.761341 19.761341 19.761341
#> 3  21.198759 21.198759 21.198759 21.198759 21.198759 21.198759 21.198759
#> 4  21.411710 21.411710 21.411710 21.411710 21.411710 21.411710 21.411710
#> 5  19.867816 19.867816 19.867816 19.867816 19.867816 19.867816 19.867816
#> 6  19.867816 19.867816 19.867816 19.867816 19.867816 19.867816 19.867816
#> 7  10.231788 10.231788 10.231788 10.231788 10.231788 10.231788 10.231788
#> 8   9.305451  9.305451  9.305451  9.305451  9.305451  9.305451  9.305451
#> 9   9.726029  9.726029  9.726029  9.726029  9.726029  9.726029  9.726029
#> 10 26.469294 26.469294 26.469294 26.469294 26.469294 26.469294 26.469294
#> 11 29.583701 29.583701 29.583701 29.583701 29.583701 29.583701 29.583701
#> 12 28.412471 28.412471 28.412471 28.412471 28.412471 28.412471 28.412471
#> 13 25.058494 25.058494 25.058494 25.058494 25.058494 25.058494 25.058494
#> 14 17.738307 17.738307 17.738307 17.738307 17.738307 17.738307 17.738307
#> 15 17.711688 17.711688 17.711688 17.711688 17.711688 17.711688 17.711688
#> 16 27.880094 27.880094 27.880094 27.880094 27.880094 27.880094 27.880094
#> 17 26.788721 26.788721 26.788721 26.788721 26.788721 26.788721 26.788721
#> 18 30.126726 30.126726 30.126726 30.126726 30.126726 30.126726 30.126726
#> 19 23.434744 23.434744 23.434744 23.434744 23.434744 23.434744 23.434744
#> 20 19.175726 19.175726 19.175726 19.175726 19.175726 19.175726 19.175726
#> 21 23.381506 23.381506 23.381506 23.381506 23.381506 23.381506 23.381506
#>       1.2081    1.1267    1.0508    0.9799    0.9139    0.8523    0.7949
#> 1  19.798585 20.042714 20.139746 20.230229 20.314613 20.393310 20.466704
#> 2  19.894855 20.112065 20.306643 20.488089 20.657305 20.815118 20.962294
#> 3  21.313143 21.341114 21.420090 21.493736 21.562419 21.626473 21.686209
#> 4  21.708708 21.548857 21.624945 21.695897 21.762068 21.823778 21.881330
#> 5  19.879528 19.696269 19.601097 19.512347 19.429580 19.352390 19.280403
#> 6  19.917471 19.696269 19.601097 19.512347 19.429580 19.352390 19.280403
#> 7  10.258247 10.295860 10.331406 10.364553 10.395466 10.424296 10.451183
#> 8   9.325155  9.392174  9.440287  9.485152  9.526994  9.566015  9.602407
#> 9   9.718911  9.802468  9.844875  9.884420  9.921300  9.955695  9.987771
#> 10 26.531028 26.655997 26.759577 26.856166 26.946246 27.030255 27.108601
#> 11 29.574053 29.521028 29.486257 29.453833 29.423595 29.395394 29.369094
#> 12 28.494342 28.551659 28.628879 28.700887 28.768042 28.830671 28.889080
#> 13 25.159498 25.279694 25.402412 25.516848 25.623572 25.723103 25.815926
#> 14 17.574995 17.618830 17.552546 17.490735 17.433091 17.379331 17.329194
#> 15 17.652182 17.939308 18.065589 18.183346 18.293168 18.395588 18.491105
#> 16 27.900661 28.032300 28.116741 28.195484 28.268920 28.337407 28.401277
#> 17 26.674128 26.794390 26.797535 26.800468 26.803203 26.805754 26.808132
#> 18 30.012662 30.050774 30.008637 29.969344 29.932699 29.898524 29.866652
#> 19 23.256443 22.829533 22.493769 22.180666 21.888665 21.616345 21.362377
#> 20 18.955973 18.328210 17.858019 17.419561 17.010654 16.629306 16.273659
#> 21 23.399433 23.470488 23.519854 23.565888 23.608820 23.648858 23.686198
#>       0.7413    0.6913    0.6447    0.6013   0.5608   0.5230   0.4877   0.4548
#> 1  20.535151 20.598984 20.658516 20.714035 20.07272 19.65403 19.39902 19.16127
#> 2  21.099551 21.227557 21.346936 21.458269 21.26118 21.29210 21.33760 21.38001
#> 3  21.741920 21.793876 21.842330 21.887519 20.93254 20.95367 21.03861 21.11789
#> 4  21.935003 21.985059 22.031740 22.075276 21.07011 21.55045 21.97203 22.36522
#> 5  19.213268 19.150657 19.092266 19.037810 17.55823 17.47887 17.51851 17.55563
#> 6  19.213268 19.150657 19.092266 19.037810 17.55823 17.57515 17.68467 17.78696
#> 7  10.476258 10.499643 10.521451 10.541790 11.33311 11.48032 11.53442 11.58478
#> 8   9.636346  9.667998  9.697517  9.725046 10.73467 10.88293 10.92336 10.96096
#> 9  10.017685 10.045584 10.071602 10.095866 11.00638 11.07831 11.06990 11.06196
#> 10 27.181668 27.249810 27.313359 27.372626 28.66248 28.88504 28.96283 29.03523
#> 11 29.344567 29.321692 29.300359 29.280464 29.41722 29.46448 29.48767 29.50930
#> 12 28.943551 28.994351 29.041728 29.085912 29.91782 30.15334 30.27078 30.38020
#> 13 25.902492 25.983225 26.058516 26.128733 24.68327 24.52772 24.52032 24.51354
#> 14 17.282437 17.238830 17.198163 17.160237 16.18251 15.70082 15.41527 15.14912
#> 15 18.580184 18.663260 18.740737 18.812992 18.67980 18.32812 18.08818 17.86443
#> 16 28.460843 28.516395 28.568202 28.616518 29.57389 29.66430 29.66812 29.67158
#> 17 26.810351 26.812420 26.814350 26.816149 27.61159 27.44741 27.27363 27.11150
#> 18 29.836928 29.809208 29.783356 29.759246 29.76802 29.53967 29.37126 29.21422
#> 19 21.125526 20.904638 20.698638 20.506521 20.41586 20.42371 20.41927 20.41522
#> 20 15.941981 15.632658 15.344182 15.075148 15.14993 15.27013 15.32929 15.38455
#> 21 23.721022 23.753498 23.783785 23.812032 25.41044 25.64942 25.71524 25.77644
#>      0.4242   0.3956   0.3689   0.3441   0.3209   0.2993   0.2791   0.2603
#> 1  18.93954 18.73343 18.65677 18.65656 18.65647 18.65642 18.65641 18.65640
#> 2  21.41956 21.45650 21.46978 21.46968 21.46963 21.46960 21.46958 21.46958
#> 3  21.19182 21.26066 21.28586 21.28587 21.28589 21.28590 21.28591 21.28592
#> 4  22.73191 23.07300 23.19908 23.19930 23.19941 23.19947 23.19950 23.19952
#> 5  17.59025 17.62239 17.63425 17.63436 17.63443 17.63448 17.63451 17.63454
#> 6  17.88235 17.97102 18.00382 18.00397 18.00406 18.00412 18.00416 18.00418
#> 7  11.63175 11.67545 11.69170 11.69169 11.69168 11.69167 11.69166 11.69165
#> 8  10.99603 11.02866 11.04082 11.04080 11.04077 11.04075 11.04073 11.04072
#> 9  11.05456 11.04769 11.04521 11.04516 11.04513 11.04510 11.04508 11.04507
#> 10 29.10274 29.16558 29.18887 29.18883 29.18879 29.18876 29.18874 29.18872
#> 11 29.52946 29.54819 29.55521 29.55525 29.55526 29.55527 29.55527 29.55528
#> 12 30.48224 30.57718 30.61234 30.61235 30.61234 30.61233 30.61231 30.61230
#> 13 24.50721 24.50144 24.49885 24.49878 24.49876 24.49876 24.49876 24.49877
#> 14 14.90090 14.67002 14.58445 14.58434 14.58430 14.58429 14.58429 14.58430
#> 15 17.65576 17.46179 17.38967 17.38944 17.38933 17.38927 17.38925 17.38923
#> 16 29.67480 29.67785 29.67897 29.67890 29.67885 29.67882 29.67880 29.67878
#> 17 26.96030 26.81967 26.76769 26.76757 26.76751 26.76747 26.76745 26.76743
#> 18 29.06777 28.93155 28.88115 28.88108 28.88105 28.88103 28.88103 28.88103
#> 19 20.41144 20.40766 20.40689 20.40712 20.40724 20.40730 20.40734 20.40736
#> 20 15.43608 15.48365 15.50217 15.50253 15.50272 15.50281 15.50286 15.50289
#> 21 25.83352 25.88662 25.90645 25.90643 25.90640 25.90638 25.90635 25.90633
#>      0.2427   0.2264   0.2111   0.1969   0.1836   0.1712   0.1597   0.1489
#> 1  18.56627 18.13982 18.13986 18.13989 19.40584 19.40615 19.40643 19.40669
#> 2  21.17172 20.28757 20.28746 20.28736 20.35863 20.35871 20.35879 20.35885
#> 3  21.29532 22.16663 22.16698 22.16727 21.05002 21.04985 21.04969 21.04955
#> 4  23.21533 23.65869 23.65874 23.65877 24.56622 24.56637 24.56651 24.56664
#> 5  17.75290 18.32673 18.32685 18.32694 18.50264 18.50262 18.50260 18.50259
#> 6  18.10630 18.55492 18.55497 18.55500 19.09405 19.09409 19.09413 19.09416
#> 7  11.65431 11.49054 11.49049 11.49044 11.28299 11.28296 11.28294 11.28292
#> 8  11.05644 11.16370 11.16373 11.16376 10.86141 10.86138 10.86135 10.86133
#> 9  11.15176 11.48125 11.48134 11.48143 11.22207 11.22204 11.22201 11.22198
#> 10 29.18132 28.99471 28.99466 28.99461 29.26230 29.26239 29.26246 29.26253
#> 11 29.80496 29.48244 29.48218 29.48195 29.83436 29.83428 29.83421 29.83415
#> 12 30.60311 30.09853 30.09833 30.09816 30.93558 30.93573 30.93586 30.93599
#> 13 24.45156 23.51514 23.51489 23.51467 21.69507 21.69459 21.69416 21.69376
#> 14 14.75584 14.49091 14.49081 14.49073 14.20520 14.20505 14.20490 14.20478
#> 15 17.32070 17.22165 17.22180 17.22194 18.25610 18.25641 18.25668 18.25693
#> 16 29.65427 29.32513 29.32505 29.32498 29.36099 29.36101 29.36103 29.36106
#> 17 26.93558 28.69729 28.69796 28.69853 28.72684 28.72701 28.72717 28.72731
#> 18 28.70891 29.18839 29.18869 29.18895 28.88551 28.88559 28.88566 28.88573
#> 19 20.25372 19.89307 19.89282 19.89261 19.17804 19.17784 19.17766 19.17750
#> 20 15.38302 14.80856 14.80811 14.80773 15.29619 15.29618 15.29617 15.29616
#> 21 25.97665 26.01434 26.01430 26.01428 25.01995 25.01975 25.01957 25.01941
#>      0.1389   0.1295   0.1208   0.1127   0.1051   0.0980   0.0914   0.0852
#> 1  19.40692 19.40714 19.40734 19.40751 19.40768 19.24214 19.24224 19.24233
#> 2  20.35892 20.35897 20.35902 20.35907 20.35911 20.54236 20.54243 20.54250
#> 3  21.04942 21.04930 21.04919 21.04910 21.04901 21.08140 21.08146 21.08151
#> 4  24.56675 24.56686 24.56696 24.56705 24.56713 24.19680 24.19659 24.19641
#> 5  18.50257 18.50256 18.50255 18.50254 18.50253 18.90772 18.90799 18.90824
#> 6  19.09419 19.09422 19.09425 19.09427 19.09429 19.41396 19.41416 19.41435
#> 7  11.28290 11.28288 11.28286 11.28285 11.28283 11.19742 11.19733 11.19724
#> 8  10.86131 10.86129 10.86127 10.86125 10.86123 10.85254 10.85254 10.85255
#> 9  11.22196 11.22193 11.22191 11.22189 11.22188 11.27024 11.27032 11.27039
#> 10 29.26259 29.26265 29.26270 29.26275 29.26280 29.35146 29.35154 29.35161
#> 11 29.83409 29.83404 29.83399 29.83395 29.83391 29.85083 29.85070 29.85058
#> 12 30.93610 30.93620 30.93629 30.93638 30.93646 30.87759 30.87753 30.87748
#> 13 21.69340 21.69307 21.69277 21.69250 21.69225 21.42666 21.42613 21.42564
#> 14 14.20466 14.20455 14.20445 14.20436 14.20428 14.17634 14.17625 14.17616
#> 15 18.25716 18.25737 18.25756 18.25774 18.25790 18.19176 18.19198 18.19218
#> 16 29.36108 29.36109 29.36111 29.36112 29.36114 29.45892 29.45897 29.45901
#> 17 28.72743 28.72755 28.72766 28.72775 28.72784 28.32351 28.32369 28.32385
#> 18 28.88579 28.88585 28.88590 28.88594 28.88598 29.22510 29.22550 29.22586
#> 19 19.17735 19.17721 19.17709 19.17697 19.17687 18.82498 18.82455 18.82416
#> 20 15.29614 15.29614 15.29613 15.29612 15.29611 15.32367 15.32347 15.32329
#> 21 25.01926 25.01912 25.01899 25.01888 25.01878 25.26459 25.26464 25.26468
#>      0.0795   0.0741   0.0691   0.0645   0.0601   0.0561   0.0523   0.0488
#> 1  19.24242 19.29769 19.29789 19.29806 19.29821 19.29833 19.29844 19.29852
#> 2  20.54256 20.67034 20.67044 20.67052 20.67058 20.67063 20.67067 20.67070
#> 3  21.08155 21.24711 21.24667 21.24629 21.24595 21.24565 21.24539 21.24517
#> 4  24.19624 23.80425 23.80462 23.80495 23.80525 23.80552 23.80577 23.80599
#> 5  18.90846 19.14051 19.14013 19.13979 19.13948 19.13920 19.13895 19.13872
#> 6  19.41452 19.62610 19.62588 19.62568 19.62550 19.62533 19.62517 19.62503
#> 7  11.19717 11.40955 11.40962 11.40968 11.40974 11.40978 11.40982 11.40985
#> 8  10.85255 10.92287 10.92281 10.92276 10.92271 10.92267 10.92264 10.92261
#> 9  11.27045 11.11315 11.11300 11.11287 11.11277 11.11268 11.11260 11.11254
#> 10 29.35167 29.38510 29.38514 29.38517 29.38520 29.38523 29.38525 29.38526
#> 11 29.85047 30.15411 30.15424 30.15434 30.15443 30.15451 30.15457 30.15463
#> 12 30.87744 30.88508 30.88539 30.88567 30.88591 30.88612 30.88630 30.88646
#> 13 21.42518 21.18812 21.18833 21.18852 21.18871 21.18888 21.18904 21.18919
#> 14 14.17608 13.83440 13.83440 13.83441 13.83442 13.83444 13.83446 13.83449
#> 15 18.19236 18.24803 18.24805 18.24806 18.24806 18.24806 18.24806 18.24805
#> 16 29.45905 29.52398 29.52400 29.52402 29.52403 29.52404 29.52405 29.52405
#> 17 28.32400 28.30452 28.30413 28.30380 28.30352 28.30327 28.30306 28.30288
#> 18 29.22619 28.91682 28.91642 28.91606 28.91575 28.91547 28.91523 28.91502
#> 19 18.82380 19.42500 19.42529 19.42554 19.42575 19.42594 19.42609 19.42623
#> 20 15.32312 14.84078 14.84125 14.84168 14.84207 14.84241 14.84271 14.84298
#> 21 25.26472 25.06251 25.06231 25.06213 25.06198 25.06185 25.06174 25.06165
#>      0.0455   0.0424   0.0396   0.0369   0.0344   0.0321   0.0299   0.0279
#> 1  19.29860 19.28642 19.28636 19.28630 19.28624 19.28618 19.28612 19.28607
#> 2  20.67072 20.32492 20.32483 20.32473 20.32464 20.32455 20.32446 20.32438
#> 3  21.24497 21.52402 21.52414 21.52425 21.52437 21.52448 21.52459 21.52469
#> 4  23.80619 23.87434 23.87434 23.87434 23.87434 23.87433 23.87433 23.87433
#> 5  19.13852 18.90059 18.90042 18.90026 18.90010 18.89994 18.89979 18.89964
#> 6  19.62491 19.40778 19.40761 19.40745 19.40729 19.40714 19.40699 19.40684
#> 7  11.40987 12.02133 12.02155 12.02177 12.02198 12.02218 12.02238 12.02258
#> 8  10.92258 11.07288 11.07293 11.07299 11.07304 11.07309 11.07314 11.07319
#> 9  11.11248 10.88626 10.88617 10.88609 10.88600 10.88592 10.88584 10.88576
#> 10 29.38527 29.00916 29.00901 29.00886 29.00871 29.00857 29.00844 29.00830
#> 11 30.15467 30.82891 30.82911 30.82931 30.82950 30.82968 30.82987 30.83004
#> 12 30.88660 30.78946 30.78939 30.78933 30.78927 30.78922 30.78916 30.78910
#> 13 21.18933 21.19757 21.19782 21.19807 21.19832 21.19856 21.19879 21.19902
#> 14 13.83451 13.55212 13.55208 13.55204 13.55201 13.55197 13.55193 13.55189
#> 15 18.24804 18.45171 18.45169 18.45166 18.45164 18.45162 18.45159 18.45157
#> 16 29.52405 29.41453 29.41448 29.41444 29.41440 29.41435 29.41431 29.41427
#> 17 28.30273 28.03809 28.03795 28.03782 28.03769 28.03757 28.03745 28.03733
#> 18 28.91483 29.47167 29.47175 29.47184 29.47192 29.47200 29.47208 29.47215
#> 19 19.42634 19.14652 19.14662 19.14671 19.14681 19.14690 19.14699 19.14708
#> 20 14.84322 14.94171 14.94176 14.94181 14.94186 14.94191 14.94195 14.94200
#> 21 25.06157 24.86001 24.85996 24.85992 24.85988 24.85984 24.85981 24.85977
#>      0.0260   0.0243   0.0226   0.0211   0.0197   0.0184   0.0171   0.0160
#> 1  19.28601 19.28596 19.28591 19.28586 19.28581 19.28576 19.28571 19.28567
#> 2  20.32429 20.32421 20.32413 20.32405 20.32398 20.32390 20.32383 20.32376
#> 3  21.52480 21.52490 21.52500 21.52509 21.52519 21.52528 21.52537 21.52545
#> 4  23.87432 23.87432 23.87432 23.87432 23.87431 23.87431 23.87431 23.87431
#> 5  18.89950 18.89936 18.89922 18.89909 18.89895 18.89883 18.89870 18.89858
#> 6  19.40670 19.40656 19.40643 19.40629 19.40616 19.40604 19.40592 19.40580
#> 7  12.02277 12.02295 12.02313 12.02331 12.02348 12.02365 12.02381 12.02397
#> 8  11.07324 11.07328 11.07333 11.07337 11.07341 11.07345 11.07350 11.07354
#> 9  10.88569 10.88561 10.88554 10.88547 10.88540 10.88533 10.88526 10.88520
#> 10 29.00817 29.00804 29.00792 29.00779 29.00768 29.00756 29.00745 29.00734
#> 11 30.83021 30.83038 30.83055 30.83071 30.83086 30.83102 30.83116 30.83131
#> 12 30.78905 30.78900 30.78895 30.78890 30.78885 30.78880 30.78876 30.78871
#> 13 21.19924 21.19946 21.19967 21.19987 21.20007 21.20027 21.20046 21.20065
#> 14 13.55186 13.55182 13.55179 13.55176 13.55173 13.55170 13.55166 13.55164
#> 15 18.45155 18.45153 18.45151 18.45149 18.45147 18.45145 18.45143 18.45141
#> 16 29.41423 29.41420 29.41416 29.41412 29.41409 29.41405 29.41402 29.41399
#> 17 28.03721 28.03710 28.03699 28.03688 28.03677 28.03667 28.03657 28.03647
#> 18 29.47223 29.47230 29.47237 29.47244 29.47251 29.47257 29.47264 29.47270
#> 19 19.14716 19.14724 19.14733 19.14741 19.14748 19.14756 19.14763 19.14770
#> 20 14.94204 14.94209 14.94213 14.94217 14.94221 14.94225 14.94228 14.94232
#> 21 24.85973 24.85970 24.85966 24.85963 24.85959 24.85956 24.85953 24.85950
#>      0.0149   0.0139   0.0130   0.0121   0.0113   0.0105   0.0098   0.0091
#> 1  19.28562 19.28558 19.28554 19.28549 19.28545 19.28542 19.28538 19.28534
#> 2  20.32369 20.32363 20.32356 20.32350 20.32344 20.32338 20.32332 20.32326
#> 3  21.52554 21.52562 21.52570 21.52578 21.52586 21.52593 21.52601 21.52608
#> 4  23.87430 23.87430 23.87430 23.87430 23.87429 23.87429 23.87429 23.87429
#> 5  18.89846 18.89835 18.89823 18.89812 18.89802 18.89791 18.89781 18.89771
#> 6  19.40568 19.40556 19.40545 19.40534 19.40524 19.40514 19.40503 19.40494
#> 7  12.02413 12.02428 12.02443 12.02457 12.02471 12.02485 12.02498 12.02512
#> 8  11.07357 11.07361 11.07365 11.07368 11.07372 11.07375 11.07379 11.07382
#> 9  10.88514 10.88508 10.88502 10.88496 10.88490 10.88485 10.88479 10.88474
#> 10 29.00723 29.00712 29.00702 29.00692 29.00682 29.00673 29.00664 29.00654
#> 11 30.83145 30.83159 30.83172 30.83185 30.83198 30.83211 30.83223 30.83235
#> 12 30.78867 30.78862 30.78858 30.78854 30.78850 30.78846 30.78843 30.78839
#> 13 21.20083 21.20100 21.20118 21.20135 21.20151 21.20167 21.20183 21.20198
#> 14 13.55161 13.55158 13.55155 13.55152 13.55150 13.55147 13.55145 13.55142
#> 15 18.45140 18.45138 18.45136 18.45135 18.45133 18.45131 18.45130 18.45128
#> 16 29.41395 29.41392 29.41389 29.41386 29.41383 29.41380 29.41378 29.41375
#> 17 28.03638 28.03628 28.03619 28.03610 28.03602 28.03593 28.03585 28.03577
#> 18 29.47276 29.47282 29.47288 29.47293 29.47299 29.47304 29.47309 29.47315
#> 19 19.14777 19.14784 19.14791 19.14797 19.14804 19.14810 19.14816 19.14822
#> 20 14.94236 14.94239 14.94243 14.94246 14.94249 14.94252 14.94255 14.94258
#> 21 24.85947 24.85944 24.85941 24.85939 24.85936 24.85933 24.85931 24.85928
#>      0.0085   0.0079   0.0074   0.0069   0.0064   0.0060
#> 1  19.28530 19.28527 19.28523 19.28520 19.28516 19.28513
#> 2  20.32320 20.32315 20.32309 20.32304 20.32299 20.32294
#> 3  21.52615 21.52622 21.52628 21.52635 21.52641 21.52647
#> 4  23.87429 23.87428 23.87428 23.87428 23.87428 23.87428
#> 5  18.89761 18.89752 18.89742 18.89733 18.89725 18.89716
#> 6  19.40484 19.40475 19.40466 19.40457 19.40448 19.40439
#> 7  12.02524 12.02537 12.02549 12.02561 12.02572 12.02584
#> 8  11.07385 11.07388 11.07391 11.07394 11.07397 11.07400
#> 9  10.88469 10.88464 10.88459 10.88454 10.88449 10.88445
#> 10 29.00646 29.00637 29.00629 29.00620 29.00612 29.00605
#> 11 30.83246 30.83257 30.83269 30.83279 30.83290 30.83300
#> 12 30.78835 30.78832 30.78828 30.78825 30.78822 30.78818
#> 13 21.20213 21.20228 21.20242 21.20256 21.20269 21.20282
#> 14 13.55140 13.55138 13.55135 13.55133 13.55131 13.55129
#> 15 18.45127 18.45125 18.45124 18.45123 18.45121 18.45120
#> 16 29.41372 29.41370 29.41367 29.41365 29.41363 29.41360
#> 17 28.03569 28.03562 28.03554 28.03547 28.03540 28.03533
#> 18 29.47320 29.47324 29.47329 29.47334 29.47338 29.47343
#> 19 19.14827 19.14833 19.14839 19.14844 19.14849 19.14854
#> 20 14.94261 14.94264 14.94267 14.94270 14.94272 14.94275
#> 21 24.85926 24.85923 24.85921 24.85919 24.85917 24.85914
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 18.7 18.1 24.4 22.8 19.2 17.8 10.4 10.4 14.7 32.4 30.4 33.9 21.5 13.3 19.2
#> [16] 27.3 26.0 30.4 19.7 15.0 21.4
#> 
#> $X
#>               am       carb        cyl        disp         drat       gear
#>  [1,] -0.8660254 -0.4524094  1.2865350  1.04894500 -1.053798227 -1.1066057
#>  [2,] -0.8660254 -1.0112680  0.1608169  0.06533137 -1.793987548 -1.1066057
#>  [3,] -0.8660254 -0.4524094 -0.9649013 -0.50516453 -0.028920706  0.2603778
#>  [4,] -0.8660254 -0.4524094 -0.9649013 -0.54815209  0.407601201  0.2603778
#>  [5,] -0.8660254  0.6653079  0.1608169 -0.35288657  0.407601201  0.2603778
#>  [6,] -0.8660254  0.6653079  0.1608169 -0.35288657  0.407601201  0.2603778
#>  [7,] -0.8660254  0.6653079  1.2865350  1.86498001 -1.471340921 -1.1066057
#>  [8,] -0.8660254  0.6653079  1.2865350  1.77754769 -1.338486427 -1.1066057
#>  [9,] -0.8660254  0.6653079  1.2865350  1.63182715 -0.901964520 -1.1066057
#> [10,]  1.1547005 -1.0112680 -0.9649013 -1.00061436  0.711268615  0.2603778
#> [11,]  1.1547005 -0.4524094 -0.9649013 -1.02247244  2.324501749  0.2603778
#> [12,]  1.1547005 -1.0112680 -0.9649013 -1.05598816  0.976977601  0.2603778
#> [13,] -0.8660254 -1.0112680 -0.9649013 -0.69897284 -0.009941493 -1.1066057
#> [14,] -0.8660254  0.6653079  1.2865350  0.97608473  0.046996147 -1.1066057
#> [15,] -0.8660254 -0.4524094  1.2865350  1.34038607 -1.186652720 -1.1066057
#> [16,]  1.1547005 -1.0112680 -0.9649013 -0.99842855  0.711268615  0.2603778
#> [17,]  1.1547005 -0.4524094 -0.9649013 -0.69751564  1.375541082  1.6273614
#> [18,]  1.1547005 -0.4524094 -0.9649013 -0.88112352  0.122913001  1.6273614
#> [19,]  1.1547005  1.7830251  0.1608169 -0.51755077 -0.161775200  1.6273614
#> [20,]  1.1547005  2.9007423  1.2865350  0.61906942 -0.313608906  1.6273614
#> [21,]  1.1547005 -0.4524094 -0.9649013 -0.69241542  0.768206255  0.2603778
#>               hp        qsec         vs          wt
#>  [1,]  0.4935733 -0.62686204 -1.1547005  0.23581402
#>  [2,] -0.4699761  1.07860878  0.8660254  0.25352221
#>  [3,] -1.0618708  0.96135766  0.8660254  0.01446169
#>  [4,] -0.6076260  2.50694059  0.8660254 -0.02095469
#>  [5,] -0.2222063  0.05532629  0.8660254  0.23581402
#>  [6,] -0.2222063  0.37510206  0.8660254  0.23581402
#>  [7,]  0.9065230 -0.11522080 -1.1547005  1.83840494
#>  [8,]  1.0441729 -0.20049434 -1.1547005  1.99246616
#>  [9,]  1.2506478 -0.41367819 -1.1547005  1.92251882
#> [10,] -1.0068108  0.67888905  0.8660254 -0.86209357
#> [11,] -1.1995207  0.17257740  0.8660254 -1.38005803
#> [12,] -1.0205758  0.90806170  0.8660254 -1.18526798
#> [13,] -0.5800961  0.96668726  0.8660254 -0.62746009
#> [14,]  1.4571227 -1.48492705 -1.1547005  0.58997776
#> [15,]  0.4935733 -0.61087325 -1.1547005  0.59440481
#> [16,] -1.0068108  0.37510206  0.8660254 -1.09672704
#> [17,] -0.6626860 -0.79740912 -1.1547005 -0.91521813
#> [18,] -0.3598562 -0.69081720  0.8660254 -1.47036978
#> [19,]  0.4935733 -1.43696068 -1.1547005 -0.35741024
#> [20,]  2.6959719 -1.91662435 -1.1547005  0.35091724
#> [21,] -0.4149162  0.21521418  0.8660254 -0.34855615
#> attr(,"center")
#>  [1]   0.4285714   2.8095238   5.7142857 216.0333333   3.7052381   3.8095238
#>  [7] 139.1428571  18.1961905   0.5714286   3.1736667
#> attr(,"scale")
#>  [1]   0.4948717   1.7893614   1.7766436 137.2490136   0.5268922   0.7315377
#>  [7]  72.6480660   1.8763147   0.4948717   1.1294211
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 32
#> 
#> $lambda.min
#> [1] 0.6913237
#> 
#> $null.dev
#> [1] 44.90884
#> 
#> $Bias
#> [1] 6.02847
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 11.18564