Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 44.25391 41.36240 36.80896 32.06635 28.06437 24.96771 22.63258 20.93564
#>   [9] 19.77126 19.04908 18.69191 18.63395 18.81920 19.20010 19.73636 20.31996
#>  [17] 20.52553 20.49867 20.12672 19.89866 19.69294 19.34719 19.00595 18.75680
#>  [25] 17.81143 15.98942 15.52213 15.32621 15.27161 15.44087 17.29209 18.10855
#>  [33] 17.96096 17.86086 17.84071 17.72921 18.62598 18.93942 18.75202 19.30909
#>  [41] 19.62204 19.46627 19.47200 19.38569 19.37635 19.05094 18.77518 18.65556
#>  [49] 22.35687 24.09367 24.09353 24.26068 63.01046 63.20644 63.05903 66.72212
#>  [57] 66.72802 66.58081 66.58633 67.54634 71.56707 71.57364 71.95321 72.15112
#>  [65] 72.15684 72.70333 72.70835 72.71313 72.71770 72.84462 72.88409 72.88727
#>  [73] 74.51142 74.51478 74.51800 74.52108 74.57762 74.58039 75.22048 75.22410
#>  [81] 75.22764 75.23108 75.23444 75.24135 75.24449 75.24758 75.25518 75.25815
#>  [89] 75.26105 75.26393 75.25692 75.25958 75.26212 75.38190 75.31841 75.32182
#>  [97] 75.35119 75.35482 75.35840 75.36193
#> 
#> $cvse
#>   [1] 12.662053 12.412299 11.773175 10.689055  9.626852  8.711113  7.904969
#>   [8]  7.185243  6.540145  5.967142  5.471048  5.062026  4.752964  4.555754
#>  [15]  4.476873  4.505376  4.623903  4.671053  4.591743  4.554723  4.522716
#>  [22]  4.485438  4.454305  4.428429  4.357541  4.321941  4.245098  4.219903
#>  [29]  4.220705  4.223737  4.733280  4.929433  4.924248  4.924155  4.929940
#>  [36]  4.940183  5.176926  5.138307  5.189305  5.200140  5.322541  5.305987
#>  [43]  5.408572  5.399063  5.402252  5.380993  5.281976  5.300112  6.802946
#>  [50]  6.682125  6.681425  6.711309 40.158270 40.147430 40.158442 40.507545
#>  [57] 40.510776 40.522461 40.525549 40.454420 40.463207 40.466100 40.454131
#>  [64] 40.446188 40.448916 40.458923 40.461548 40.464124 40.466654 40.463379
#>  [71] 40.556651 40.559095 40.617976 40.620358 40.622690 40.624973 40.623590
#>  [78] 40.625783 40.728778 40.731035 40.733253 40.735431 40.737572 40.739422
#>  [85] 40.741492 40.743524 40.745221 40.747181 40.749106 40.750993 40.753526
#>  [92] 40.755356 40.757157 40.752365 40.757401 40.759058 40.761044 40.762619
#>  [99] 40.764165 40.765683
#> 
#> $fold
#>  [1]  7  9  3 10  6  5  6  8  5  9  2  2  3  1  1  4  4 10  8  7  1
#> 
#> $lambda
#>   [1] 5.411498624 5.046781729 4.706645531 4.389433375 4.093600256 3.817705300
#>   [7] 3.560404740 3.320445377 3.096658472 2.887954055 2.693315617 2.511795159
#>  [13] 2.342508572 2.184631334 2.037394494 1.900080924 1.772021829 1.652593489
#>  [19] 1.541214218 1.437341538 1.340469529 1.250126369 1.165872036 1.087296163
#>  [25] 1.014016041 0.945674753 0.881939440 0.822499673 0.767065948 0.715368271
#>  [31] 0.667154843 0.622190840 0.580257260 0.541149862 0.504678173 0.470664553
#>  [37] 0.438943338 0.409360026 0.381770530 0.356040474 0.332044538 0.309665847
#>  [43] 0.288795405 0.269331562 0.251179516 0.234250857 0.218463133 0.203739449
#>  [49] 0.190008092 0.177202183 0.165259349 0.154121422 0.143734154 0.134046953
#>  [55] 0.125012637 0.116587204 0.108729616 0.101401604 0.094567475 0.088193944
#>  [61] 0.082249967 0.076706595 0.071536827 0.066715484 0.062219084 0.058025726
#>  [67] 0.054114986 0.050467817 0.047066455 0.043894334 0.040936003 0.038177053
#>  [73] 0.035604047 0.033204454 0.030966585 0.028879541 0.026933156 0.025117952
#>  [79] 0.023425086 0.021846313 0.020373945 0.019000809 0.017720218 0.016525935
#>  [85] 0.015412142 0.014373415 0.013404695 0.012501264 0.011658720 0.010872962
#>  [91] 0.010140160 0.009456748 0.008819394 0.008224997 0.007670659 0.007153683
#>  [97] 0.006671548 0.006221908 0.005802573 0.005411499
#> 
#> $fit
#> $beta
#>               5.4115     5.0468     4.7066     4.3894   4.0936    3.8177
#> (Intercept) 20.97143 22.8468698 24.6669412 26.3643458 27.94735 29.423667
#> am           0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#> cyl          0.00000 -0.3029559 -0.5969674 -0.8711636 -1.12688 -1.365362
#> disp         0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#> wt           0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#>                3.5604   3.3204   3.0967    2.8880    2.6933    2.5118    2.3425
#> (Intercept) 30.800484 32.08451 33.28199 34.398773 35.440285 36.411602 37.317456
#> am           0.000000  0.00000  0.00000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.00000  0.00000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.587771 -1.79519 -1.98863 -2.169033 -2.337277 -2.494182 -2.640512
#> disp         0.000000  0.00000  0.00000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.00000  0.00000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.00000  0.00000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.00000  0.00000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.00000  0.00000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.00000  0.00000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.00000  0.00000  0.000000  0.000000  0.000000  0.000000
#>               2.1846    2.0374    1.9001    1.7720    1.6526    1.5412
#> (Intercept) 38.16226 38.950124 39.684890 40.225301 40.225301 40.225301
#> am           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.77698 -2.904251 -3.022944 -3.110241 -3.110241 -3.110241
#> disp         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.4373    1.3405    1.2501    1.1659    1.0873    1.0140
#> (Intercept) 40.225301 40.225301 40.225301 40.225301 40.225301 40.798330
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -3.110241 -3.110241 -3.110241 -3.110241 -3.110241 -2.009313
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000 -2.343082
#>                0.9457    0.8819    0.8225     0.7671     0.7154    0.6672
#> (Intercept) 40.958382 40.958067 40.957861 40.9746040 41.0001617 41.024033
#> am           0.000000  0.000000  0.000000  0.0000000  0.0000000  0.000000
#> carb         0.000000  0.000000  0.000000 -0.0611222 -0.1474408 -0.227734
#> cyl         -1.699310 -1.699134 -1.699018 -1.6778355 -1.6489910 -1.622215
#> disp         0.000000  0.000000  0.000000  0.0000000  0.0000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.0000000  0.0000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.0000000  0.0000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.0000000  0.0000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.0000000  0.0000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.0000000  0.0000000  0.000000
#> wt          -3.002441 -3.002688 -3.002849 -2.9989784 -2.9920161 -2.985463
#>                 0.6222     0.5803     0.5411     0.5047     0.4707     0.4389
#> (Intercept) 41.0463005 41.0666194 41.0859718 41.1040599 41.1209333 41.1366701
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb        -0.3026126 -0.3723882 -0.4375425 -0.4982835 -0.5549282 -0.6077548
#> cyl         -1.5972473 -1.5737402 -1.5520022 -1.5317628 -1.5128913 -1.4952920
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.9793480 -2.9739864 -2.9686834 -2.9637032 -2.9590547 -2.9547192
#>                 0.4094    0.3818    0.3560    0.3320    0.3097    0.2888
#> (Intercept) 41.1513462 41.165033 30.931048 30.930874 30.930862 30.930927
#> am           0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb        -0.6570211 -0.702967 -1.077924 -1.077991 -1.078031 -1.078054
#> cyl         -1.4788789 -1.463572 -1.084100 -1.083938 -1.083835 -1.083770
#> disp         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.0000000  0.000000  2.014253  2.014242  2.014214  2.014181
#> gear         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.0000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.9506758 -2.946905 -2.422379 -2.422572 -2.422706 -2.422798
#>                0.2693    0.2512    0.2343    0.2185    0.2037    0.1900
#> (Intercept) 30.931020 30.931114 30.931199 30.931271 35.548155 35.546694
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb        -1.078067 -1.078073 -1.078076 -1.078077 -0.721594 -0.721526
#> cyl         -1.083730 -1.083706 -1.083692 -1.083684 -1.243575 -1.243444
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         2.014150  2.014123  2.014101  2.014084  2.405105  2.405819
#> gear         0.000000  0.000000  0.000000  0.000000 -1.231575 -1.232014
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.422860 -2.422901 -2.422928 -2.422946 -2.862663 -2.862808
#>                 0.1772     0.1653     0.1541     0.1437     0.1340     0.1250
#> (Intercept) 35.5454729 36.9004398 36.8995128 36.8990832 36.8990109 36.8991827
#> am           0.0000000  1.6910659  1.6904402  1.6900040  1.6897186  1.6895501
#> carb        -0.7214999 -0.7395436 -0.7396872 -0.7397963 -0.7398764 -0.7399327
#> cyl         -1.2433296 -1.2916248 -1.2913705 -1.2911642 -1.2910012 -1.2908764
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         2.4063164  2.0835855  2.0833738  2.0831530  2.0829376  2.0827371
#> gear        -1.2322859 -1.6909215 -1.6904212 -1.6900539 -1.6897953 -1.6896233
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.8629117 -2.5059118 -2.5062615 -2.5065625 -2.5068141 -2.5070187
#>                 0.1166    0.1087     0.1014     0.0946     0.0882     0.0822
#> (Intercept) 36.8995092 36.899922 36.9003707 27.5680323 27.5628311 27.5581302
#> am           1.6894692  1.689452  1.6894773  2.3974849  2.3982751  2.3989893
#> carb        -0.7399702 -0.739993 -0.7400046 -0.6494070 -0.6493845 -0.6493642
#> cyl         -1.2907839 -1.290718 -1.2906740 -0.8330658 -0.8327626 -0.8324885
#> disp         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         2.0825572  2.082401  2.0822683  1.9463458  1.9461567  1.9459858
#> gear        -1.6895187 -1.689465 -1.6894478 -1.4614304 -1.4613451 -1.4612678
#> hp           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0000000  0.000000  0.0000000  0.3650635  0.3652706  0.3654578
#> vs           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.5071807 -2.507305 -2.5073976 -2.7970718 -2.7972014 -2.7973185
#>                 0.0767     0.0715     0.0667     0.0622     0.0580     0.0541
#> (Intercept) 27.5538816 27.5500419 27.5465717 27.5434355 27.5406012 27.5380397
#> am           2.3996347  2.4002180  2.4007451  2.4012214  2.4016519  2.4020410
#> carb        -0.6493459 -0.6493293 -0.6493144 -0.6493009 -0.6492887 -0.6492777
#> cyl         -0.8322408 -0.8320170 -0.8318147 -0.8316319 -0.8314667 -0.8313173
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         1.9458313  1.9456916  1.9455654  1.9454513  1.9453482  1.9452550
#> gear        -1.4611980 -1.4611347 -1.4610775 -1.4610258 -1.4609790 -1.4609367
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.3656270  0.3657799  0.3659181  0.3660429  0.3661558  0.3662578
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.7974243 -2.7975199 -2.7976063 -2.7976843 -2.7977548 -2.7978185
#>                   0.0505       0.0471       0.0439     0.0409       0.0382
#> (Intercept) 26.443471060 26.437491891 26.432124117 26.4273187 26.423029656
#> am           2.377637891  2.378574818  2.379434480  2.3802217  2.380941134
#> carb        -0.648423294 -0.648390029 -0.648359714 -0.6483321 -0.648307134
#> cyl         -0.629811415 -0.629509381 -0.629243336 -0.6290100 -0.628806394
#> disp         0.000000000  0.000000000  0.000000000  0.0000000  0.000000000
#> drat         1.957572050  1.957329282  1.957108263  1.9569075  1.956725481
#> gear        -1.207939228 -1.207932427 -1.207937802 -1.2079535 -1.207977958
#> hp          -0.007684741 -0.007681588 -0.007678503 -0.0076755 -0.007672591
#> qsec         0.335699725  0.335971068  0.336217498  0.3364408  0.336642679
#> vs           0.000000000  0.000000000  0.000000000  0.0000000  0.000000000
#> wt          -2.634427958 -2.634689972 -2.634928577 -2.6351454 -2.635341983
#>                   0.0356       0.0332       0.0310      0.0289       0.0269
#> (Intercept) 26.419213662 26.415830278 26.412841688 26.41021264 26.407910308
#> am           2.381597279  2.382194446  2.382736760  2.38322816  2.383672382
#> carb        -0.648284490 -0.648264040 -0.648245616 -0.64822906 -0.648214229
#> cyl         -0.628629653 -0.628477190 -0.628346601 -0.62823567 -0.628142344
#> disp         0.000000000  0.000000000  0.000000000  0.00000000  0.000000000
#> drat         1.956560895  1.956412414  1.956278799  1.95615888  1.956051567
#> gear        -1.208009640 -1.208047265 -1.208089678 -1.20813585 -1.208184889
#> hp          -0.007669782 -0.007667081 -0.007664493 -0.00766202 -0.007659664
#> qsec         0.336824743  0.336988520  0.337135451  0.33726689  0.337384122
#> vs           0.000000000  0.000000000  0.000000000  0.00000000  0.000000000
#> wt          -2.635519812 -2.635680287 -2.635824736 -2.63595442 -2.636070507
#>                   0.0251       0.0234       0.0218       0.0204       0.0190
#> (Intercept) 26.405904232 26.404166159 26.402669951 26.401391473 26.400308479
#> am           2.384072989  2.384433338  2.384756603  2.385045774  2.385303661
#> carb        -0.648200977 -0.648189174 -0.648178698 -0.648169433 -0.648161273
#> cyl         -0.628064758 -0.628001188 -0.627950060 -0.627909938 -0.627879510
#> disp         0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> drat         1.955955813  1.955870654  1.955795179  1.955728541  1.955669948
#> gear        -1.208235995 -1.208288479 -1.208341741 -1.208395264 -1.208448603
#> hp          -0.007657427 -0.007655307 -0.007653303 -0.007651414 -0.007649637
#> qsec         0.337488334  0.337580649  0.337662111  0.337733696  0.337796309
#> vs           0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> wt          -2.636174124 -2.636266311 -2.636348045 -2.636420239 -2.636483746
#>                   0.0177       0.0165       0.0154        0.0144       0.0134
#> (Intercept) 26.399400505 26.398648761 26.398036031  2.640129e+01 26.749021963
#> am           2.385532896  2.385735946  2.385915112  2.385481e+00  2.355968603
#> carb        -0.648154116 -0.648147870 -0.648142449 -6.480088e-01 -0.717617483
#> cyl         -0.627857586 -0.627843084 -0.627835026 -6.277383e-01 -0.606908972
#> disp         0.000000000  0.000000000  0.000000000 -9.078908e-06 -0.002320894
#> drat         1.955618665  1.955574008  1.955535343  1.954822e+00  2.007806044
#> gear        -1.208501381 -1.208553280 -1.208604036 -1.208777e+00 -1.236545670
#> hp          -0.007647969 -0.007646407 -0.007644947 -7.636416e-03 -0.005790078
#> qsec         0.337850792  0.337897924  0.337938426  3.379633e-01  0.301450939
#> vs           0.000000000  0.000000000  0.000000000  0.000000e+00  0.000000000
#> wt          -2.636539359 -2.636587812 -2.636629790 -2.636745e+00 -2.459704376
#>                   0.0125       0.0117       0.0109       0.0101       0.0095
#> (Intercept) 26.744375910 26.739859511 26.735474907 26.731223761 26.727107284
#> am           2.356088895  2.356220009  2.356360848  2.356510366  2.356667568
#> carb        -0.717616798 -0.717609089 -0.717594881 -0.717574681 -0.717548972
#> cyl         -0.606565112 -0.606235883 -0.605921071 -0.605620441 -0.605333738
#> disp        -0.002319775 -0.002318528 -0.002317167 -0.002315703 -0.002314146
#> drat         2.007788122  2.007760336  2.007723464  2.007678253  2.007625416
#> gear        -1.236145240 -1.235763859 -1.235401133 -1.235056650 -1.234729981
#> hp          -0.005796301 -0.005802304 -0.005808087 -0.005813652 -0.005818999
#> qsec         0.301558195  0.301667361  0.301778023  0.301889794  0.302002309
#> vs           0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> wt          -2.459781683 -2.459875235 -2.459983626 -2.460105510 -2.460239603
#>                   0.0088       0.0082       0.0077       0.0072       0.0067
#> (Intercept) 26.697474820 26.696629747 26.695588729 26.694406087 26.693120338
#> am           2.371796898  2.372100860  2.372361786  2.372588391  2.372788724
#> carb        -0.714948994 -0.714875212 -0.714794545 -0.714709934 -0.714623136
#> cyl         -0.590623135 -0.590420058 -0.590234020 -0.590060986 -0.589898487
#> disp        -0.002274529 -0.002271260 -0.002268055 -0.002264913 -0.002261832
#> drat         2.008870983  2.008784767  2.008694171  2.008600161  2.008503356
#> gear        -1.228650755 -1.228434850 -1.228220954 -1.228011578 -1.227808760
#> hp          -0.006017420 -0.006026825 -0.006035456 -0.006043391 -0.006050704
#> qsec         0.293430375  0.293378104  0.293347143  0.293333125  0.293332926
#> vs           0.080723472  0.081214687  0.081639364  0.082010678  0.082337928
#> wt          -2.448927749 -2.448903946 -2.448918963 -2.448968340 -2.449047568
#>                   0.0062       0.0058       0.0054
#> (Intercept) 26.691759314 26.690343898 26.688890539
#> am           2.372969752  2.373137214  2.373295648
#> carb        -0.714535160 -0.714446562 -0.714357656
#> cyl         -0.589745035 -0.589599739 -0.589462060
#> disp        -0.002258810 -0.002255841 -0.002252925
#> drat         2.008404166  2.008302879  2.008199719
#> gear        -1.227614034 -1.227428484 -1.227252815
#> hp          -0.006057459 -0.006063711 -0.006069508
#> qsec         0.293344235  0.293365265  0.293394566
#> vs           0.082628016  0.082886310  0.083117146
#> wt          -2.449152361 -2.449278788 -2.449423345
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1   1 266  18   1   1   6   6   6   6   5   5   5   5
#>  [37]   5   5   5  37   1   1   1   1   1   1   1  28   1   1  40   1   1   1
#>  [55]   1   1   1   1  74   1   1   1   1   1   1   1   1  66   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   2  83   1   1   1
#>  [91]   1   1  21   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.411498624 5.046781729 4.706645531 4.389433375 4.093600256 3.817705300
#>   [7] 3.560404740 3.320445377 3.096658472 2.887954055 2.693315617 2.511795159
#>  [13] 2.342508572 2.184631334 2.037394494 1.900080924 1.772021829 1.652593489
#>  [19] 1.541214218 1.437341538 1.340469529 1.250126369 1.165872036 1.087296163
#>  [25] 1.014016041 0.945674753 0.881939440 0.822499673 0.767065948 0.715368271
#>  [31] 0.667154843 0.622190840 0.580257260 0.541149862 0.504678173 0.470664553
#>  [37] 0.438943338 0.409360026 0.381770530 0.356040474 0.332044538 0.309665847
#>  [43] 0.288795405 0.269331562 0.251179516 0.234250857 0.218463133 0.203739449
#>  [49] 0.190008092 0.177202183 0.165259349 0.154121422 0.143734154 0.134046953
#>  [55] 0.125012637 0.116587204 0.108729616 0.101401604 0.094567475 0.088193944
#>  [61] 0.082249967 0.076706595 0.071536827 0.066715484 0.062219084 0.058025726
#>  [67] 0.054114986 0.050467817 0.047066455 0.043894334 0.040936003 0.038177053
#>  [73] 0.035604047 0.033204454 0.030966585 0.028879541 0.026933156 0.025117952
#>  [79] 0.023425086 0.021846313 0.020373945 0.019000809 0.017720218 0.016525935
#>  [85] 0.015412142 0.014373415 0.013404695 0.012501264 0.011658720 0.010872962
#>  [91] 0.010140160 0.009456748 0.008819394 0.008224997 0.007670659 0.007153683
#>  [97] 0.006671548 0.006221908 0.005802573 0.005411499
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 24
#> 
#> $loss
#>   [1] 815.6629 702.2928 603.3691 520.9650 452.6839 396.4578 350.5035 313.2843
#>   [9] 283.4771 259.9438 241.7060 227.9236 217.8761 210.9463 206.6062 204.4048
#>  [17] 203.9229 203.9229 203.9229 203.9229 203.9229 203.9229 203.9229 203.9229
#>  [25] 141.1846 137.9999 137.9999 137.9999 135.3026 131.8773 129.0944 126.8495
#>  [33] 125.0620 123.6580 122.5797 121.7746 121.1981 120.8122 120.5841 110.6727
#>  [41] 110.6727 110.6727 110.6727 110.6727 110.6727 110.6727 110.6727 106.3632
#>  [49] 106.3632 106.3631 103.3489 103.3489 103.3489 103.3489 103.3489 103.3489
#>  [57] 103.3489 103.3489 102.0119 102.0119 102.0119 102.0119 102.0119 102.0118
#>  [65] 102.0118 102.0118 102.0118 101.5052 101.5052 101.5052 101.5052 101.5052
#>  [73] 101.5052 101.5052 101.5052 101.5052 101.5052 101.5052 101.5051 101.5051
#>  [81] 101.5051 101.5051 101.5051 101.5051 101.5051 101.5044 101.4180 101.4180
#>  [89] 101.4180 101.4180 101.4180 101.4179 101.4095 101.4095 101.4095 101.4095
#>  [97] 101.4095 101.4095 101.4095 101.4095
#> 
#> $linear.predictors
#>      5.4115   5.0468   4.7066   4.3894   4.0936   3.8177   3.5604   3.3204
#> 1  20.97143 21.02913 21.08514 21.13736 21.18607 21.23150 21.27386 21.31337
#> 2  20.97143 21.63505 22.27907 22.87969 23.43983 23.96222 24.44940 24.90375
#> 3  20.97143 21.02913 21.08514 21.13736 21.18607 21.23150 21.27386 21.31337
#> 4  20.97143 20.42322 19.89120 19.39504 18.93231 18.50077 18.09832 17.72299
#> 5  20.97143 21.02913 21.08514 21.13736 21.18607 21.23150 21.27386 21.31337
#> 6  20.97143 20.42322 19.89120 19.39504 18.93231 18.50077 18.09832 17.72299
#> 7  20.97143 21.63505 22.27907 22.87969 23.43983 23.96222 24.44940 24.90375
#> 8  20.97143 21.02913 21.08514 21.13736 21.18607 21.23150 21.27386 21.31337
#> 9  20.97143 20.42322 19.89120 19.39504 18.93231 18.50077 18.09832 17.72299
#> 10 20.97143 20.42322 19.89120 19.39504 18.93231 18.50077 18.09832 17.72299
#> 11 20.97143 20.42322 19.89120 19.39504 18.93231 18.50077 18.09832 17.72299
#> 12 20.97143 20.42322 19.89120 19.39504 18.93231 18.50077 18.09832 17.72299
#> 13 20.97143 21.63505 22.27907 22.87969 23.43983 23.96222 24.44940 24.90375
#> 14 20.97143 21.63505 22.27907 22.87969 23.43983 23.96222 24.44940 24.90375
#> 15 20.97143 21.63505 22.27907 22.87969 23.43983 23.96222 24.44940 24.90375
#> 16 20.97143 20.42322 19.89120 19.39504 18.93231 18.50077 18.09832 17.72299
#> 17 20.97143 20.42322 19.89120 19.39504 18.93231 18.50077 18.09832 17.72299
#> 18 20.97143 21.63505 22.27907 22.87969 23.43983 23.96222 24.44940 24.90375
#> 19 20.97143 21.63505 22.27907 22.87969 23.43983 23.96222 24.44940 24.90375
#> 20 20.97143 20.42322 19.89120 19.39504 18.93231 18.50077 18.09832 17.72299
#> 21 20.97143 21.02913 21.08514 21.13736 21.18607 21.23150 21.27386 21.31337
#>      3.0967   2.8880   2.6933   2.5118   2.3425   2.1846   2.0374   1.9001
#> 1  21.35022 21.38458 21.41662 21.44651 21.47438 21.50038 21.52462 21.54723
#> 2  25.32747 25.72264 26.09118 26.43487 26.75541 27.05434 27.33312 27.59311
#> 3  21.35022 21.38458 21.41662 21.44651 21.47438 21.50038 21.52462 21.54723
#> 4  17.37296 17.04651 16.74207 16.45815 16.19336 15.94642 15.71612 15.50134
#> 5  21.35022 21.38458 21.41662 21.44651 21.47438 21.50038 21.52462 21.54723
#> 6  17.37296 17.04651 16.74207 16.45815 16.19336 15.94642 15.71612 15.50134
#> 7  25.32747 25.72264 26.09118 26.43487 26.75541 27.05434 27.33312 27.59311
#> 8  21.35022 21.38458 21.41662 21.44651 21.47438 21.50038 21.52462 21.54723
#> 9  17.37296 17.04651 16.74207 16.45815 16.19336 15.94642 15.71612 15.50134
#> 10 17.37296 17.04651 16.74207 16.45815 16.19336 15.94642 15.71612 15.50134
#> 11 17.37296 17.04651 16.74207 16.45815 16.19336 15.94642 15.71612 15.50134
#> 12 17.37296 17.04651 16.74207 16.45815 16.19336 15.94642 15.71612 15.50134
#> 13 25.32747 25.72264 26.09118 26.43487 26.75541 27.05434 27.33312 27.59311
#> 14 25.32747 25.72264 26.09118 26.43487 26.75541 27.05434 27.33312 27.59311
#> 15 25.32747 25.72264 26.09118 26.43487 26.75541 27.05434 27.33312 27.59311
#> 16 17.37296 17.04651 16.74207 16.45815 16.19336 15.94642 15.71612 15.50134
#> 17 17.37296 17.04651 16.74207 16.45815 16.19336 15.94642 15.71612 15.50134
#> 18 25.32747 25.72264 26.09118 26.43487 26.75541 27.05434 27.33312 27.59311
#> 19 25.32747 25.72264 26.09118 26.43487 26.75541 27.05434 27.33312 27.59311
#> 20 17.37296 17.04651 16.74207 16.45815 16.19336 15.94642 15.71612 15.50134
#> 21 21.35022 21.38458 21.41662 21.44651 21.47438 21.50038 21.52462 21.54723
#>      1.7720   1.6526   1.5412   1.4373   1.3405   1.2501   1.1659   1.0873
#> 1  21.56386 21.56386 21.56386 21.56386 21.56386 21.56386 21.56386 21.56386
#> 2  27.78434 27.78434 27.78434 27.78434 27.78434 27.78434 27.78434 27.78434
#> 3  21.56386 21.56386 21.56386 21.56386 21.56386 21.56386 21.56386 21.56386
#> 4  15.34337 15.34337 15.34337 15.34337 15.34337 15.34337 15.34337 15.34337
#> 5  21.56386 21.56386 21.56386 21.56386 21.56386 21.56386 21.56386 21.56386
#> 6  15.34337 15.34337 15.34337 15.34337 15.34337 15.34337 15.34337 15.34337
#> 7  27.78434 27.78434 27.78434 27.78434 27.78434 27.78434 27.78434 27.78434
#> 8  21.56386 21.56386 21.56386 21.56386 21.56386 21.56386 21.56386 21.56386
#> 9  15.34337 15.34337 15.34337 15.34337 15.34337 15.34337 15.34337 15.34337
#> 10 15.34337 15.34337 15.34337 15.34337 15.34337 15.34337 15.34337 15.34337
#> 11 15.34337 15.34337 15.34337 15.34337 15.34337 15.34337 15.34337 15.34337
#> 12 15.34337 15.34337 15.34337 15.34337 15.34337 15.34337 15.34337 15.34337
#> 13 27.78434 27.78434 27.78434 27.78434 27.78434 27.78434 27.78434 27.78434
#> 14 27.78434 27.78434 27.78434 27.78434 27.78434 27.78434 27.78434 27.78434
#> 15 27.78434 27.78434 27.78434 27.78434 27.78434 27.78434 27.78434 27.78434
#> 16 15.34337 15.34337 15.34337 15.34337 15.34337 15.34337 15.34337 15.34337
#> 17 15.34337 15.34337 15.34337 15.34337 15.34337 15.34337 15.34337 15.34337
#> 18 27.78434 27.78434 27.78434 27.78434 27.78434 27.78434 27.78434 27.78434
#> 19 27.78434 27.78434 27.78434 27.78434 27.78434 27.78434 27.78434 27.78434
#> 20 15.34337 15.34337 15.34337 15.34337 15.34337 15.34337 15.34337 15.34337
#> 21 21.56386 21.56386 21.56386 21.56386 21.56386 21.56386 21.56386 21.56386
#>      1.0140   0.9457   0.8819   0.8225   0.7671   0.7154   0.6672   0.6222
#> 1  22.60358 22.89612 22.89622 22.89628 22.80578 22.67737 22.55789 22.44647
#> 2  27.32513 27.19548 27.19530 27.19518 27.24451 27.31528 27.38117 27.44261
#> 3  21.20944 21.10967 21.10962 21.10959 21.20475 21.33944 21.46475 21.58160
#> 4  16.66362 17.03550 17.03575 17.03591 17.11319 17.22082 17.32085 17.41414
#> 5  20.63539 20.37407 20.37396 20.37389 20.47000 20.60640 20.73331 20.85166
#> 6  16.35902 16.64518 16.64540 16.64554 16.60108 16.53697 16.47728 16.42160
#> 7  25.38037 24.70345 24.70306 24.70281 24.69424 24.68447 24.67550 24.66714
#> 8  20.68225 20.43412 20.43402 20.43395 20.34662 20.22392 20.10981 20.00341
#> 9  15.18748 15.14396 15.14406 15.14412 15.16271 15.18841 15.21228 15.23454
#> 10 15.98413 16.16479 16.16497 16.16509 16.18236 16.20569 16.22733 16.24752
#> 11 12.42264 11.60108 11.60088 11.60075 11.56279 11.51039 11.46170 11.41629
#> 12 12.20005 11.31585 11.31563 11.31548 11.27789 11.22614 11.17808 11.13326
#> 13 27.60630 27.55577 27.55562 27.55552 27.60439 27.67432 27.73942 27.80013
#> 14 28.97700 29.31220 29.31219 29.31219 29.29767 29.27721 29.25818 29.24044
#> 15 28.46152 28.65166 28.65160 28.65156 28.69901 28.76641 28.82912 28.88760
#> 16 16.47618 16.79530 16.79553 16.79568 16.87327 16.98146 17.08202 17.17579
#> 17 15.71467 15.81951 15.81966 15.81976 15.89860 16.00905 16.11174 16.20750
#> 18 27.74688 27.73592 27.73578 27.73569 27.72320 27.70640 27.69081 27.67628
#> 19 29.21599 29.61845 29.61846 29.61848 29.60356 29.58240 29.56270 29.54433
#> 20 17.29625 17.84616 17.84647 17.84668 17.80067 17.73378 17.67146 17.61334
#> 21 22.25211 22.44576 22.44582 22.44586 22.23369 21.93369 21.65461 21.39435
#>      0.5803   0.5411   0.5047   0.4707   0.4389   0.4094   0.3818   0.3560
#> 1  22.34278 22.24584 22.15545 22.07115 21.99253 21.91922 21.85084 21.62371
#> 2  27.49962 27.55307 27.60293 27.64943 27.69280 27.73324 27.77096 27.65168
#> 3  21.69042 21.79210 21.88689 21.97530 22.05774 22.13463 22.20633 21.76448
#> 4  17.50141 17.58260 17.65825 17.72880 17.79459 17.85595 17.91317 18.11432
#> 5  20.96180 21.06477 21.16079 21.25033 21.33383 21.41171 21.48434 20.52643
#> 6  16.37001 16.32158 16.27640 16.23427 16.19497 16.15832 16.12414 15.76442
#> 7  24.65883 24.65153 24.64478 24.63849 24.63263 24.62716 24.62206 24.70418
#> 8  19.90411 19.81152 19.72521 19.64472 19.56966 19.49966 19.43438 19.67764
#> 9  15.25541 15.27478 15.29283 15.30967 15.32536 15.34000 15.35365 15.34916
#> 10 16.26656 16.28414 16.30049 16.31574 16.32997 16.34323 16.35560 16.17276
#> 11 11.37372 11.33420 11.29738 11.26305 11.23104 11.20118 11.17334 11.13083
#> 12 11.09119 11.05217 11.01583 10.98194 10.95034 10.92087 10.89338 11.50498
#> 13 27.85650 27.90932 27.95858 28.00452 28.04736 28.08732 28.12459 28.40564
#> 14 29.22389 29.20845 29.19406 29.18064 29.16812 29.15645 29.14556 30.45693
#> 15 28.94201 28.99289 29.04033 29.08457 29.12584 29.16432 29.20021 29.57181
#> 16 17.26349 17.34510 17.42116 17.49207 17.55821 17.61989 17.67742 17.13497
#> 17 16.29694 16.38028 16.45795 16.53038 16.59793 16.66092 16.71967 16.99226
#> 18 27.66255 27.64990 27.63812 27.62713 27.61689 27.60734 27.59843 28.17805
#> 19 29.52724 29.51126 29.49636 29.48246 29.46950 29.45742 29.44614 28.36748
#> 20 17.55961 17.50906 17.46188 17.41789 17.37685 17.33859 17.30290 18.76776
#> 21 21.15191 20.92545 20.71432 20.51744 20.33382 20.16257 20.00287 18.54051
#>      0.3320   0.3097   0.2888   0.2693   0.2512   0.2343   0.2185   0.2037
#> 1  21.62369 21.62367 21.62367 21.62367 21.62367 21.62367 21.62367 22.15376
#> 2  27.65159 27.65154 27.65150 27.65148 27.65146 27.65145 27.65145 27.54423
#> 3  21.76455 21.76460 21.76464 21.76466 21.76468 21.76469 21.76470 21.87465
#> 4  18.11460 18.11479 18.11491 18.11498 18.11503 18.11506 18.11507 18.19016
#> 5  20.52646 20.52649 20.52651 20.52653 20.52655 20.52656 20.52657 20.40366
#> 6  15.76454 15.76463 15.76468 15.76473 15.76475 15.76477 15.76478 16.51913
#> 7  24.70387 24.70365 24.70352 24.70343 24.70337 24.70333 24.70331 24.61499
#> 8  19.67746 19.67734 19.67726 19.67720 19.67717 19.67715 19.67713 19.85448
#> 9  15.34925 15.34931 15.34935 15.34938 15.34940 15.34941 15.34941 15.47268
#> 10 16.17292 16.17303 16.17311 16.17315 16.17318 16.17320 16.17322 16.44598
#> 11 11.13063 11.13050 11.13041 11.13036 11.13032 11.13030 11.13029 11.03643
#> 12 11.50476 11.50461 11.50450 11.50443 11.50439 11.50436 11.50434 11.48601
#> 13 28.40558 28.40553 28.40550 28.40547 28.40546 28.40545 28.40544 28.44093
#> 14 30.45690 30.45686 30.45683 30.45681 30.45679 30.45677 30.45676 31.43833
#> 15 29.57181 29.57181 29.57180 29.57180 29.57179 29.57179 29.57179 29.82251
#> 16 17.13524 17.13543 17.13555 17.13563 17.13569 17.13572 17.13575 17.02315
#> 17 16.99246 16.99259 16.99268 16.99273 16.99276 16.99278 16.99279 16.86242
#> 18 28.17793 28.17784 28.17777 28.17773 28.17770 28.17768 28.17767 27.50131
#> 19 28.36748 28.36749 28.36751 28.36753 28.36754 28.36755 28.36756 27.70883
#> 20 18.76795 18.76806 18.76813 18.76816 18.76818 18.76818 18.76819 17.63020
#> 21 18.54033 18.54023 18.54017 18.54014 18.54013 18.54013 18.54013 18.37617
#>      0.1900   0.1772   0.1653   0.1541   0.1437   0.1340   0.1250   0.1166
#> 1  22.15401 22.15416 22.68039 22.68005 22.67980 22.67964 22.67954 22.67948
#> 2  27.54402 27.54387 28.12987 28.12956 28.12933 28.12917 28.12906 28.12898
#> 3  21.87446 21.87433 21.69932 21.69950 21.69965 21.69978 21.69989 21.69997
#> 4  18.19032 18.19046 17.95855 17.95900 17.95937 17.95968 17.95991 17.96010
#> 5  20.40321 20.40290 20.41862 20.41879 20.41894 20.41907 20.41919 20.41929
#> 6  16.51945 16.51966 16.27871 16.27881 16.27892 16.27901 16.27910 16.27917
#> 7  24.61477 24.61460 23.76520 23.76507 23.76491 23.76472 23.76454 23.76436
#> 8  19.85462 19.85470 18.97615 18.97614 18.97608 18.97599 18.97589 18.97577
#> 9  15.47276 15.47282 15.47359 15.47370 15.47379 15.47387 15.47394 15.47400
#> 10 16.44611 16.44621 16.32560 16.32583 16.32602 16.32619 16.32633 16.32644
#> 11 11.03630 11.03620 11.48537 11.48495 11.48461 11.48434 11.48414 11.48400
#> 12 11.48608 11.48612 11.87239 11.87187 11.87143 11.87108 11.87080 11.87058
#> 13 28.44090 28.44088 28.90980 28.90949 28.90925 28.90906 28.90893 28.90883
#> 14 31.43906 31.43955 31.40726 31.40683 31.40647 31.40617 31.40593 31.40574
#> 15 29.82264 29.82272 30.11616 30.11595 30.11578 30.11566 30.11557 30.11551
#> 16 17.02302 17.02296 16.94548 16.94598 16.94642 16.94678 16.94708 16.94732
#> 17 16.86247 16.86254 16.79780 16.79813 16.79839 16.79861 16.79878 16.79891
#> 18 27.50116 27.50108 27.35894 27.35894 27.35889 27.35883 27.35875 27.35867
#> 19 27.70830 27.70795 27.55498 27.55534 27.55563 27.55586 27.55604 27.55619
#> 20 17.63042 17.63063 17.69472 17.69512 17.69542 17.69563 17.69578 17.69586
#> 21 18.37589 18.37567 18.55109 18.55097 18.55089 18.55085 18.55083 18.55083
#>      0.1087   0.1014   0.0946   0.0882   0.0822   0.0767   0.0715   0.0667
#> 1  22.67946 22.67946 22.79514 22.79531 22.79547 22.79561 22.79573 22.79585
#> 2  28.12894 28.12893 27.93618 27.93617 27.93617 27.93616 27.93615 27.93615
#> 3  21.70003 21.70008 21.63493 21.63486 21.63479 21.63473 21.63467 21.63462
#> 4  17.96023 17.96032 17.94284 17.94285 17.94286 17.94287 17.94288 17.94288
#> 5  20.41938 20.41944 20.61157 20.61168 20.61179 20.61188 20.61197 20.61204
#> 6  16.27924 16.27929 15.96642 15.96620 15.96600 15.96582 15.96566 15.96552
#> 7  23.76420 23.76406 24.27009 24.27008 24.27007 24.27007 24.27006 24.27005
#> 8  18.97566 18.97556 18.81470 18.81435 18.81404 18.81375 18.81350 18.81327
#> 9  15.47404 15.47408 15.51430 15.51434 15.51438 15.51442 15.51445 15.51448
#> 10 16.32653 16.32659 16.53831 16.53844 16.53856 16.53867 16.53876 16.53885
#> 11 11.48389 11.48383 11.50359 11.50365 11.50371 11.50376 11.50380 11.50384
#> 12 11.87042 11.87030 11.61734 11.61722 11.61710 11.61700 11.61691 11.61683
#> 13 28.90877 28.90874 29.03344 29.03359 29.03371 29.03383 29.03394 29.03403
#> 14 31.40559 31.40549 31.32791 31.32779 31.32768 31.32759 31.32750 31.32743
#> 15 30.11547 30.11545 30.48384 30.48409 30.48432 30.48453 30.48471 30.48488
#> 16 16.94751 16.94765 16.90524 16.90528 16.90532 16.90536 16.90539 16.90542
#> 17 16.79900 16.79907 16.68474 16.68471 16.68469 16.68467 16.68466 16.68464
#> 18 27.35859 27.35852 26.76043 26.76004 26.75970 26.75939 26.75911 26.75885
#> 19 27.55629 27.55636 27.30261 27.30248 27.30236 27.30225 27.30215 27.30206
#> 20 17.69591 17.69592 18.03649 18.03682 18.03711 18.03738 18.03762 18.03784
#> 21 18.55084 18.55086 18.71989 18.72003 18.72016 18.72027 18.72037 18.72046
#>      0.0622   0.0580   0.0541   0.0505   0.0471   0.0439   0.0409   0.0382
#> 1  22.79595 22.79605 22.79613 23.02942 23.02953 23.02962 23.02971 23.02978
#> 2  27.93614 27.93614 27.93614 27.77915 27.77918 27.77921 27.77925 27.77930
#> 3  21.63458 21.63453 21.63450 21.63268 21.63260 21.63252 21.63245 21.63239
#> 4  17.94289 17.94290 17.94290 17.95701 17.95704 17.95707 17.95708 17.95709
#> 5  20.61211 20.61218 20.61223 20.66109 20.66122 20.66133 20.66143 20.66152
#> 6  15.96539 15.96527 15.96516 15.50109 15.50104 15.50100 15.50098 15.50097
#> 7  24.27005 24.27005 24.27004 24.12833 24.12839 24.12844 24.12849 24.12854
#> 8  18.81306 18.81287 18.81270 19.04848 19.04798 19.04751 19.04707 19.04667
#> 9  15.51451 15.51453 15.51455 15.58144 15.58147 15.58150 15.58153 15.58154
#> 10 16.53893 16.53900 16.53906 16.54428 16.54446 16.54462 16.54476 16.54489
#> 11 11.50388 11.50391 11.50394 11.55292 11.55295 11.55297 11.55300 11.55303
#> 12 11.61675 11.61669 11.61662 11.50981 11.50967 11.50954 11.50944 11.50936
#> 13 29.03412 29.03420 29.03427 29.04172 29.04187 29.04201 29.04214 29.04226
#> 14 31.32735 31.32729 31.32723 31.38704 31.38687 31.38671 31.38657 31.38645
#> 15 30.48503 30.48517 30.48530 30.42938 30.42970 30.43000 30.43028 30.43053
#> 16 16.90544 16.90547 16.90549 17.12457 17.12455 17.12453 17.12450 17.12446
#> 17 16.68463 16.68461 16.68460 16.76311 16.76306 16.76301 16.76296 16.76291
#> 18 26.75862 26.75841 26.75823 26.90656 26.90601 26.90551 26.90505 26.90463
#> 19 27.30198 27.30191 27.30184 27.16443 27.16432 27.16423 27.16415 27.16408
#> 20 18.03804 18.03822 18.03838 17.89792 17.89837 17.89879 17.89916 17.89950
#> 21 18.72055 18.72062 18.72069 18.75957 18.75972 18.75986 18.75999 18.76011
#>      0.0356   0.0332   0.0310   0.0289   0.0269   0.0251   0.0234   0.0218
#> 1  23.02984 23.02989 23.02994 23.02998 23.03001 23.03003 23.03005 23.03007
#> 2  27.77934 27.77939 27.77944 27.77949 27.77954 27.77960 27.77965 27.77970
#> 3  21.63233 21.63228 21.63223 21.63218 21.63214 21.63211 21.63207 21.63204
#> 4  17.95710 17.95711 17.95711 17.95711 17.95710 17.95710 17.95709 17.95709
#> 5  20.66160 20.66167 20.66173 20.66178 20.66183 20.66187 20.66190 20.66193
#> 6  15.50097 15.50098 15.50100 15.50103 15.50106 15.50110 15.50114 15.50118
#> 7  24.12858 24.12861 24.12865 24.12867 24.12870 24.12872 24.12875 24.12876
#> 8  19.04630 19.04596 19.04564 19.04535 19.04509 19.04485 19.04463 19.04443
#> 9  15.58156 15.58157 15.58158 15.58159 15.58159 15.58159 15.58159 15.58159
#> 10 16.54500 16.54510 16.54519 16.54526 16.54533 16.54539 16.54544 16.54548
#> 11 11.55306 11.55308 11.55311 11.55313 11.55316 11.55318 11.55320 11.55322
#> 12 11.50928 11.50923 11.50918 11.50915 11.50912 11.50910 11.50909 11.50908
#> 13 29.04237 29.04247 29.04256 29.04265 29.04273 29.04280 29.04286 29.04292
#> 14 31.38633 31.38623 31.38613 31.38605 31.38598 31.38591 31.38585 31.38580
#> 15 30.43075 30.43096 30.43115 30.43132 30.43147 30.43161 30.43174 30.43185
#> 16 17.12442 17.12438 17.12434 17.12429 17.12424 17.12420 17.12415 17.12410
#> 17 16.76286 16.76282 16.76277 16.76273 16.76269 16.76266 16.76262 16.76259
#> 18 26.90425 26.90391 26.90360 26.90332 26.90306 26.90284 26.90264 26.90245
#> 19 27.16402 27.16397 27.16392 27.16388 27.16385 27.16382 27.16380 27.16378
#> 20 17.89981 17.90009 17.90034 17.90057 17.90078 17.90096 17.90112 17.90126
#> 21 18.76021 18.76030 18.76038 18.76046 18.76052 18.76058 18.76063 18.76067
#>      0.0204   0.0190   0.0177   0.0165   0.0154   0.0144   0.0134   0.0125
#> 1  23.03008 23.03009 23.03009 23.03010 23.03010 23.03015 22.98653 22.98666
#> 2  27.77975 27.77980 27.77985 27.77989 27.77994 27.77985 27.85796 27.85770
#> 3  21.63202 21.63199 21.63197 21.63195 21.63193 21.63189 21.58091 21.58092
#> 4  17.95708 17.95707 17.95706 17.95705 17.95703 17.95681 17.89399 17.89412
#> 5  20.66195 20.66197 20.66198 20.66200 20.66201 20.66244 20.67646 20.67653
#> 6  15.50122 15.50127 15.50131 15.50136 15.50141 15.50196 15.49845 15.49800
#> 7  24.12878 24.12880 24.12881 24.12882 24.12883 24.12899 24.08889 24.08893
#> 8  19.04425 19.04408 19.04393 19.04380 19.04368 19.04414 19.11552 19.11559
#> 9  15.58159 15.58158 15.58158 15.58157 15.58156 15.58228 15.74716 15.74715
#> 10 16.54552 16.54555 16.54557 16.54559 16.54561 16.54636 16.64375 16.64379
#> 11 11.55324 11.55326 11.55327 11.55329 11.55330 11.55256 11.42073 11.42076
#> 12 11.50908 11.50908 11.50909 11.50910 11.50912 11.50864 11.55010 11.54987
#> 13 29.04298 29.04303 29.04307 29.04311 29.04315 29.04296 29.09850 29.09848
#> 14 31.38576 31.38572 31.38568 31.38565 31.38562 31.38492 31.32809 31.32808
#> 15 30.43195 30.43204 30.43212 30.43219 30.43225 30.43208 30.43044 30.43049
#> 16 17.12405 17.12401 17.12396 17.12392 17.12388 17.12409 17.11118 17.11140
#> 17 16.76256 16.76253 16.76250 16.76247 16.76245 16.76187 16.67347 16.67362
#> 18 26.90229 26.90215 26.90203 26.90192 26.90182 26.90112 26.91833 26.91830
#> 19 27.16376 27.16375 27.16374 27.16374 27.16373 27.16400 27.12681 27.12670
#> 20 17.90139 17.90150 17.90160 17.90169 17.90176 17.90108 17.90002 17.90024
#> 21 18.76071 18.76075 18.76077 18.76080 18.76082 18.76182 18.75268 18.75268
#>      0.0117   0.0109   0.0101   0.0095   0.0088   0.0082   0.0077   0.0072
#> 1  22.98679 22.98692 22.98705 22.98717 22.97356 22.97375 22.97393 22.97409
#> 2  27.85745 27.85721 27.85697 27.85674 27.86606 27.86599 27.86590 27.86579
#> 3  21.58093 21.58094 21.58095 21.58097 21.60312 21.60331 21.60348 21.60364
#> 4  17.89424 17.89436 17.89449 17.89461 17.88258 17.88260 17.88263 17.88267
#> 5  20.67660 20.67666 20.67673 20.67679 20.69432 20.69444 20.69455 20.69465
#> 6  15.49758 15.49718 15.49680 15.49644 15.48738 15.48696 15.48658 15.48623
#> 7  24.08897 24.08902 24.08906 24.08911 24.05950 24.05908 24.05874 24.05848
#> 8  19.11565 19.11568 19.11570 19.11571 19.15895 19.15915 19.15933 19.15949
#> 9  15.74713 15.74710 15.74706 15.74701 15.73703 15.73681 15.73659 15.73638
#> 10 16.64382 16.64385 16.64388 16.64390 16.62835 16.62811 16.62789 16.62769
#> 11 11.42080 11.42085 11.42090 11.42097 11.42459 11.42486 11.42512 11.42535
#> 12 11.54964 11.54942 11.54920 11.54899 11.55263 11.55257 11.55249 11.55241
#> 13 29.09845 29.09842 29.09839 29.09836 29.10344 29.10346 29.10346 29.10345
#> 14 31.32807 31.32806 31.32805 31.32804 31.34096 31.34114 31.34129 31.34143
#> 15 30.43053 30.43058 30.43063 30.43069 30.42802 30.42798 30.42794 30.42791
#> 16 17.11161 17.11181 17.11200 17.11218 17.10515 17.10532 17.10547 17.10561
#> 17 16.67376 16.67390 16.67404 16.67417 16.66796 16.66813 16.66829 16.66844
#> 18 26.91826 26.91821 26.91815 26.91808 26.87130 26.87113 26.87097 26.87082
#> 19 27.12659 27.12649 27.12640 27.12631 27.14526 27.14533 27.14539 27.14545
#> 20 17.90045 17.90065 17.90084 17.90103 17.92335 17.92343 17.92350 17.92356
#> 21 18.75268 18.75269 18.75271 18.75273 18.74648 18.74646 18.74645 18.74646
#>      0.0067   0.0062   0.0058   0.0054
#> 1  22.97424 22.97438 22.97450 22.97462
#> 2  27.86567 27.86555 27.86542 27.86529
#> 3  21.60378 21.60392 21.60405 21.60417
#> 4  17.88272 17.88277 17.88283 17.88289
#> 5  20.69474 20.69482 20.69490 20.69498
#> 6  15.48592 15.48564 15.48539 15.48517
#> 7  24.05826 24.05808 24.05793 24.05781
#> 8  19.15962 19.15974 19.15983 19.15990
#> 9  15.73617 15.73597 15.73577 15.73558
#> 10 16.62751 16.62735 16.62720 16.62706
#> 11 11.42558 11.42578 11.42598 11.42616
#> 12 11.55231 11.55221 11.55210 11.55198
#> 13 29.10342 29.10339 29.10335 29.10331
#> 14 31.34155 31.34166 31.34175 31.34184
#> 15 30.42789 30.42787 30.42786 30.42785
#> 16 17.10574 17.10586 17.10598 17.10609
#> 17 16.66858 16.66872 16.66886 16.66899
#> 18 26.87068 26.87054 26.87040 26.87025
#> 19 27.14551 27.14556 27.14562 27.14567
#> 20 17.92363 17.92369 17.92376 17.92382
#> 21 18.74648 18.74650 18.74653 18.74656
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 22.8 21.4 18.7 18.1 14.3 22.8 19.2 16.4 17.3 10.4 14.7 32.4 30.4 33.9
#> [16] 15.5 19.2 26.0 30.4 15.8 19.7
#> 
#> $X
#>               am       carb        cyl        disp        drat       gear
#>  [1,]  1.1547005  1.0139612 -0.1097643 -0.58106129  0.54149693  0.3735437
#>  [2,]  1.1547005 -1.1887821 -1.2622890 -0.99364919  0.45663547  0.3735437
#>  [3,] -0.8660254 -1.1887821 -0.1097643  0.19650821 -0.85023101 -0.9338592
#>  [4,] -0.8660254 -0.4545343  1.0427605  1.00581524 -0.73142496 -0.9338592
#>  [5,] -0.8660254 -1.1887821 -0.1097643 -0.06532642 -1.39334435 -0.9338592
#>  [6,] -0.8660254  1.0139612  1.0427605  1.00581524 -0.62959121 -0.9338592
#>  [7,] -0.8660254 -0.4545343 -1.2622890 -0.73340144  0.57544152  0.3735437
#>  [8,] -0.8660254  1.0139612 -0.1097643 -0.52075998  0.57544152  0.3735437
#>  [9,] -0.8660254  0.2797134  1.0427605  0.33774022 -0.86720330 -0.9338592
#> [10,] -0.8660254  0.2797134  1.0427605  0.33774022 -0.86720330 -0.9338592
#> [11,] -0.8660254  1.0139612  1.0427605  1.89446611 -1.10481538 -0.9338592
#> [12,] -0.8660254  1.0139612  1.0427605  1.64056586 -0.59564663 -0.9338592
#> [13,]  1.1547005 -1.1887821 -1.2622890 -1.22612660  0.84699819  0.3735437
#> [14,]  1.1547005 -0.4545343 -1.2622890 -1.24992975  2.28964300  0.3735437
#> [15,]  1.1547005 -1.1887821 -1.2622890 -1.28642791  1.08461027  0.3735437
#> [16,] -0.8660254 -0.4545343  1.0427605  0.67257117 -1.39334435 -0.9338592
#> [17,] -0.8660254 -0.4545343  1.0427605  1.32319055 -0.85023101 -0.9338592
#> [18,]  1.1547005 -0.4545343 -1.2622890 -0.89605628  1.44102841  1.6809466
#> [19,]  1.1547005 -0.4545343 -1.2622890 -1.09600273  0.32085714  1.6809466
#> [20,]  1.1547005  1.0139612  1.0427605  0.93440580  1.08461027  1.6809466
#> [21,]  1.1547005  2.4824568 -0.1097643 -0.70007703  0.06627276  1.6809466
#>               hp        qsec         vs           wt
#>  [1,] -0.5512483 -0.75807668 -0.8660254 -0.529744799
#>  [2,] -0.8356351  0.42461724  1.1547005 -0.827779469
#>  [3,] -0.5512483  0.88119211  1.1547005  0.061357296
#>  [4,]  0.5361129 -0.45002617 -0.8660254  0.284883299
#>  [5,] -0.6348915  1.31026246  1.1547005  0.304752277
#>  [6,]  1.7071173 -1.09913260 -0.8660254  0.414031656
#>  [7,] -0.8021778  2.78450419  1.1547005 -0.003216882
#>  [8,] -0.3337761  0.25408928  1.1547005  0.284883299
#>  [9,]  0.6197561 -0.24099190 -0.8660254  0.910756105
#> [10,]  0.6197561 -0.13097386 -0.8660254  0.572983479
#> [11,]  1.0379719  0.07806042 -0.8660254  2.083025806
#> [12,]  1.4561878 -0.22999009 -0.8660254  2.177403452
#> [13,] -1.2873082  0.89769481  1.1547005 -0.946993336
#> [14,] -1.5215091  0.37510912  1.1547005 -1.528160942
#> [15,] -1.3040368  1.13423360  1.1547005 -1.309602185
#> [16,]  0.1178970 -0.53253970 -0.8660254  0.364359211
#> [17,]  0.5361129 -0.43352346 -0.8660254  0.687230103
#> [18,] -0.8690924 -0.62605503 -0.8660254 -1.006600270
#> [19,] -0.5010624 -0.51603699  1.1547005 -1.629492730
#> [20,]  2.0249613 -1.83625347 -0.8660254  0.016652096
#> [21,]  0.5361129 -1.28616327 -0.8660254 -0.380727464
#> attr(,"center")
#>  [1]   0.4285714   2.6190476   6.1904762 233.2333333   3.5809524   3.7142857
#>  [7] 142.9523810  17.8380952   0.4285714   3.1532381
#> attr(,"scale")
#>  [1]   0.4948717   1.3619381   1.7353207 126.0337496   0.5891956   0.7648752
#>  [7]  59.7777441   1.8178837   0.4948717   1.0065943
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 29
#> 
#> $lambda.min
#> [1] 0.7670659
#> 
#> $null.dev
#> [1] 38.84109
#> 
#> $Bias
#> [1] 10.39003
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 6.296611