Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 39.65920 37.07948 32.88016 29.16192 26.08405 23.55262 21.48665 19.81635
#>   [9] 18.48167 17.43093 16.61973 16.00995 15.56890 15.26858 15.08505 14.93143
#>  [17] 14.51057 14.24688 14.30675 14.38610 14.45222 14.50873 14.45607 14.30374
#>  [25] 14.12722 13.91845 13.09081 12.97709 12.93934 13.09195 13.03194 12.95442
#>  [33] 12.08675 11.75231 11.36818 13.02494 12.89961 13.94162 16.43770 16.45262
#>  [41] 14.20950 14.20899 14.78326 14.88199 14.83038 17.08602 17.33200 17.33203
#>  [49] 17.28491 17.06891 17.79753 18.32460 17.93417 17.66032 17.65958 19.66856
#>  [57] 19.83016 19.83092 19.62757 19.79870 19.75144 19.22153 17.66753 17.21586
#>  [65] 17.22987 17.50133 17.50107 17.66486 17.66468 17.66957 18.54811 18.54859
#>  [73] 18.52385 18.52414 18.52442 19.12995 19.13064 19.13131 19.05111 18.94790
#>  [81] 18.94870 18.94947 18.95021 18.95093 18.95163 18.95231 18.95293 18.91863
#>  [89] 18.91920 18.91975 18.92028 18.90501 18.90549 18.90595 18.90640 18.91525
#>  [97] 18.91567 18.91608 18.91648 18.91687
#> 
#> $cvse
#>   [1] 11.670741 11.561043 10.561333  9.570379  8.674991  7.863377  7.129487
#>   [8]  6.472030  5.893631  5.399948  4.998456  4.696609  4.499365  4.406658
#>  [15]  4.412000  4.459466  4.315836  4.227949  4.225216  4.220165  4.213177
#>  [22]  4.204959  4.195825  4.168034  4.136277  4.099701  4.162448  4.135133
#>  [29]  4.104240  4.150008  4.137648  4.124670  3.549528  3.387522  3.369720
#>  [36]  3.229643  3.104086  3.179695  4.057917  4.053005  3.716818  3.716725
#>  [43]  3.869928  3.916800  3.913147  4.621490  4.625663  4.625702  4.626594
#>  [50]  4.600024  4.602519  4.491724  4.517570  4.509302  4.509186  5.224335
#>  [57]  5.226276  5.226703  5.188464  5.164810  5.116004  5.157751  5.135031
#>  [64]  5.119176  5.113829  5.157744  5.157751  5.148678  5.148650  5.134737
#>  [71]  5.297019  5.297142  5.301421  5.301550  5.301672  5.242205  5.242298
#>  [78]  5.242386  5.179359  5.120566  5.120567  5.120564  5.120559  5.120551
#>  [85]  5.120541  5.120528  5.120520  5.126589  5.126582  5.126573  5.126563
#>  [92]  5.124927  5.124926  5.124924  5.124922  5.123487  5.123485  5.123483
#>  [99]  5.123481  5.123479
#> 
#> $fold
#>  [1]  1  7  8  4  8  1  6  4 10  2  7  3  9 10  3  6  2  1  9  5  5
#> 
#> $lambda
#>   [1] 5.338589466 4.978786403 4.643232863 4.330294509 4.038447152 3.766269330
#>   [7] 3.512435382 3.275708993 3.054937170 2.849044630 2.657028557 2.477953725
#>  [13] 2.310947937 2.155197781 2.009944664 1.874481120 1.748147367 1.630328085
#>  [19] 1.520449429 1.417976226 1.322409374 1.233283408 1.150164234 1.072647014
#>  [25] 1.000354195 0.932933671 0.870057064 0.811418130 0.756731263 0.705730109
#>  [31] 0.658166262 0.613808058 0.572439450 0.533858947 0.497878640 0.464323286
#>  [37] 0.433029451 0.403844715 0.376626933 0.351243538 0.327570899 0.305493717
#>  [43] 0.284904463 0.265702856 0.247795373 0.231094794 0.215519778 0.200994466
#>  [49] 0.187448112 0.174814737 0.163032808 0.152044943 0.141797623 0.132240937
#>  [55] 0.123328341 0.115016423 0.107264701 0.100035420 0.093293367 0.087005706
#>  [61] 0.081141813 0.075673126 0.070573011 0.065816626 0.061380806 0.057243945
#>  [67] 0.053385895 0.049787864 0.046432329 0.043302945 0.040384472 0.037662693
#>  [73] 0.035124354 0.032757090 0.030549372 0.028490446 0.026570286 0.024779537
#>  [79] 0.023109479 0.021551978 0.020099447 0.018744811 0.017481474 0.016303281
#>  [85] 0.015204494 0.014179762 0.013224094 0.012332834 0.011501642 0.010726470
#>  [91] 0.010003542 0.009329337 0.008700571 0.008114181 0.007567313 0.007057301
#>  [97] 0.006581663 0.006138081 0.005724394 0.005338589
#> 
#> $fit
#> $beta
#>               5.3386     4.9788    4.6432    4.3303    4.0384   3.7663
#> (Intercept) 20.71429 22.4671130 24.101806 25.626325 27.048098 28.37405
#> am           0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> carb         0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> cyl          0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> disp         0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> drat         0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> gear         0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> hp           0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> qsec         0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> vs           0.00000  0.0000000  0.000000  0.000000  0.000000  0.00000
#> wt           0.00000 -0.5557138 -1.073974 -1.557306 -2.008063 -2.42844
#>                3.5124    3.2757    3.0549    2.8490    2.6570    2.4780
#> (Intercept) 29.610632 30.763875 31.839394 32.842426 33.777857 34.650243
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.820485 -3.186107 -3.527088 -3.845088 -4.141656 -4.418236
#>                2.3109    2.1552    2.0099    1.8745    1.7481    1.6303
#> (Intercept) 35.463834 36.222591 36.930210 37.590139 38.052706 38.052706
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -4.676175 -4.916731 -5.141073 -5.350296 -5.496948 -5.496948
#>                1.5204    1.4180      1.3224      1.2333     1.1502     1.0726
#> (Intercept) 38.052706 38.052706 37.65182769 36.20527958 34.8561569 33.5979606
#> am           0.000000  0.000000  0.00000000  0.00000000  0.0000000  0.0000000
#> carb         0.000000  0.000000  0.00000000  0.00000000  0.0000000  0.0000000
#> cyl          0.000000  0.000000  0.00000000  0.00000000  0.0000000  0.0000000
#> disp         0.000000  0.000000  0.00000000  0.00000000  0.0000000  0.0000000
#> drat         0.000000  0.000000  0.00000000  0.00000000  0.0000000  0.0000000
#> gear         0.000000  0.000000  0.00000000  0.00000000  0.0000000  0.0000000
#> hp           0.000000  0.000000  0.00000000  0.00000000  0.0000000  0.0000000
#> qsec         0.000000  0.000000  0.02105964  0.09705224  0.1679267  0.2340245
#> vs           0.000000  0.000000  0.00000000  0.00000000  0.0000000  0.0000000
#> wt          -5.496948 -5.496948 -5.49048630 -5.46717036 -5.4454248 -5.4251447
#>                 1.0004     0.9329     0.8701     0.8114     0.7567     0.7057
#> (Intercept) 32.4245625 31.3302475 30.3096856 29.3579062 28.4702736 26.3568098
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.2878104
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.2956675  0.3531559  0.4067699  0.4567704  0.5034011  0.5436063
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.4062315 -5.3885930 -5.3721432 -5.3568022 -5.3424950 -5.2283677
#>                0.6582    0.6138    0.5724    0.5339     0.4979     0.4643
#> (Intercept)  9.052566  9.050477  9.049179  9.048374 17.4370027 15.3905655
#> am           4.207349  4.207890  4.208225  4.208433  5.3294753  5.3959176
#> carb         0.000000  0.000000  0.000000  0.000000 -0.9098102 -0.8385736
#> cyl          0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> qsec         1.190486  1.190553  1.190595  1.190621  0.6734692  0.7747574
#> vs           0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> wt          -3.693715 -3.693510 -3.693382 -3.693303 -2.7461732 -2.7489705
#>                 0.4330     0.4038     0.3766     0.3512     0.3276     0.3055
#> (Intercept) 15.3913822 15.3920142 15.3924936 15.3928526 15.3931192 15.3933158
#> am           5.3952116  5.3946919  5.3943108  5.3940319  5.3938282  5.3936796
#> carb        -0.8384251 -0.8383182 -0.8382410 -0.8381852 -0.8381448 -0.8381154
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.7747683  0.7747748  0.7747788  0.7747812  0.7747828  0.7747838
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.7493264 -2.7495869 -2.7497773 -2.7499162 -2.7500174 -2.7500912
#>                 0.2849     0.2657     0.2478     0.2311     0.2155     0.2010
#> (Intercept) 15.3934603 11.5181179 11.5194788 11.5204879 11.5212311 11.5217749
#> am           5.3935713  4.3439121  4.3432514  4.3427979  4.3424899  4.3422835
#> carb        -0.8380941 -0.9490510 -0.9489793 -0.9489206 -0.9488735 -0.9488361
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  1.9606051  1.9606458  1.9606345  1.9605963  1.9605467
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.7747845  0.5804879  0.5804577  0.5804426  0.5804367  0.5804361
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.7501449 -2.3869559 -2.3872335 -2.3874438 -2.3876018 -2.3877197
#>                 0.1874     0.1748     0.1630     0.1520      0.1418     0.1322
#> (Intercept) 11.5221703 14.3449749 14.3471247 14.3488631 -15.4350544 -15.439206
#> am           4.3421472  4.2442393  4.2436846  4.2432322   4.1622089   4.162344
#> carb        -0.9488069 -0.9895014 -0.9896425 -0.9897560  -1.7879382  -1.788130
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000   1.7505825   1.750878
#> disp         0.0000000  0.0000000  0.0000000  0.0000000   0.0000000   0.000000
#> drat         1.9604949  2.1261909  2.1268556  2.1273926   3.4632306   3.463356
#> gear         0.0000000  0.0000000  0.0000000  0.0000000   2.5257038   2.526134
#> hp           0.0000000  0.0000000  0.0000000  0.0000000   0.0000000   0.000000
#> qsec         0.5804385  0.3042034  0.3039351  0.3037188   0.6561209   0.656141
#> vs           0.0000000  1.3363933  1.3369196  1.3373418   2.5491674   2.549454
#> wt          -2.3878071 -2.0407178 -2.0404951 -2.0403177  -1.9048459  -1.904750
#>                  0.1233      0.1150      0.1073      0.1000      0.0933
#> (Intercept) -15.4429920 -15.4464446 -15.4495932 -15.4524646 -15.4550833
#> am            4.1624663   4.1625780   4.1626796   4.1627722   4.1628565
#> carb         -1.7883052  -1.7884651  -1.7886110  -1.7887442  -1.7888657
#> cyl           1.7511478   1.7513938   1.7516181   1.7518226   1.7520092
#> disp          0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> drat          3.4634700   3.4635742   3.4636691   3.4637557   3.4638347
#> gear          2.5265274   2.5268862   2.5272136   2.5275125   2.5277853
#> hp            0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> qsec          0.6561593   0.6561758   0.6561908   0.6562045   0.6562169
#> vs            2.5497163   2.5499553   2.5501733   2.5503722   2.5505538
#> wt           -1.9046617  -1.9045813  -1.9045078  -1.9044406  -1.9043793
#>                  0.0870      0.0811      0.0757      0.0706      0.0658
#> (Intercept) -15.4574714 -15.4596493 -15.4616356 -15.4634470 -15.4650990
#> am            4.1629332   4.1630031   4.1630668   4.1631248   4.1631777
#> carb         -1.7889767  -1.7890779  -1.7891703  -1.7892546  -1.7893316
#> cyl           1.7521794   1.7523345   1.7524761   1.7526052   1.7527229
#> disp          0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> drat          3.4639067   3.4639723   3.4640322   3.4640868   3.4641366
#> gear          2.5280342   2.5282614   2.5284687   2.5286579   2.5288306
#> hp            0.0000000   0.0000000   0.0000000   0.0000000   0.0000000
#> qsec          0.6562281   0.6562384   0.6562477   0.6562561   0.6562638
#> vs            2.5507194   2.5508705   2.5510084   2.5511341   2.5512489
#> wt           -1.9043232  -1.9042720  -1.9042252  -1.9041825  -1.9041435
#>                  0.0614      0.0572      0.0534        0.0498        0.0464
#> (Intercept) -15.4666056 -15.4679797 -15.4692328 -17.137899481 -17.132434383
#> am            4.1632258   4.1632697   4.1633096   4.245242062   4.245036850
#> carb         -1.7894018  -1.7894658  -1.7895243  -1.687516490  -1.687354214
#> cyl           1.7528303   1.7529282   1.7530175   1.654073635   1.653762827
#> disp          0.0000000   0.0000000   0.0000000   0.006150234   0.006147838
#> drat          3.4641820   3.4642233   3.4642611   3.626536237   3.626337001
#> gear          2.5289881   2.5291318   2.5292630   2.466575059   2.466142196
#> hp            0.0000000   0.0000000   0.0000000   0.000000000   0.000000000
#> qsec          0.6562708   0.6562772   0.6562829   0.761396752   0.761319324
#> vs            2.5513536   2.5514491   2.5515362   2.579542276   2.579220279
#> wt           -1.9041079  -1.9040753  -1.9040456  -2.427716510  -2.427579549
#>                    0.0433        0.0404        0.0377        0.0351
#> (Intercept) -17.127465279 -17.122958623 -17.118882405 -17.115206141
#> am            4.244841286   4.244655296   4.244478758   4.244311513
#> carb         -1.687208469  -1.687078015  -1.686961679  -1.686858354
#> cyl           1.653483844   1.653234297   1.653011930   1.652814611
#> disp          0.006145514   0.006143265   0.006141095   0.006139008
#> drat          3.626163065   3.626012257   3.625882547   3.625772037
#> gear          2.465749823   2.465395112   2.465075373   2.464788052
#> hp            0.000000000   0.000000000   0.000000000   0.000000000
#> qsec          0.761246080   0.761176923   0.761111744   0.761050424
#> vs            2.578930128   2.578669496   2.578436176   2.578228085
#> wt           -2.427442707  -2.427306779  -2.427172454  -2.427040322
#>                    0.0328        0.0305        0.0285        0.0266
#> (Intercept) -17.111900860 -17.108939080 -17.106294781 -17.103943371
#> am            4.244153364   4.244004090   4.243863443   4.243731160
#> carb         -1.686766998  -1.686686632  -1.686616333  -1.686555241
#> cyl           1.652640332   1.652487203   1.652353449   1.652237408
#> disp          0.006137005   0.006135087   0.006133254   0.006131508
#> drat          3.625678953   3.625601644   3.625538576   3.625488322
#> gear          2.464530729   2.464301116   2.464097052   2.463916501
#> hp            0.000000000   0.000000000   0.000000000   0.000000000
#> qsec          0.760992836   0.760938848   0.760888324   0.760841122
#> vs            2.578043251   2.577879818   2.577736040   2.577610274
#> wt           -2.426910887  -2.426784574  -2.426661733  -2.426542651
#>                    0.0248        0.0231        0.0216       0.0201
#> (Intercept) -17.101861649 -17.100027766 -17.098421179 -17.09702261
#> am            4.243606958   4.243490545   4.243381621   4.24327988
#> carb         -1.686502548  -1.686457499  -1.686419392  -1.68638757
#> cyl           1.652137521   1.652052334   1.651980489   1.65192072
#> disp          0.006129846   0.006128268   0.006126773   0.00612536
#> drat          3.625449563   3.625421076   3.625401733   3.62539049
#> gear          2.463757547   2.463618389   2.463497340   2.46339282
#> hp            0.000000000   0.000000000   0.000000000   0.00000000
#> qsec          0.760797102   0.760756121   0.760718036   0.76068271
#> vs            2.577500983   2.577406725   2.577326152   2.57725801
#> wt           -2.426427555  -2.426316622  -2.426209979  -2.42610771
#>                    0.0187        0.0175        0.0163        0.0152
#> (Intercept) -17.095813996 -17.094778440 -17.093900168 -17.093164472
#> am            4.243185006   4.243096695   4.243014634   4.242938516
#> carb         -1.686361426  -1.686340395  -1.686323951  -1.686311613
#> cyl           1.651871858   1.651832807   1.651802558   1.651780176
#> disp          0.006124025   0.006122767   0.006121584   0.006120473
#> drat          3.625386406   3.625388588   3.625396239   3.625408624
#> gear          2.463303346   2.463227547   2.463164134   2.463111915
#> hp            0.000000000   0.000000000   0.000000000   0.000000000
#> qsec          0.760649990   0.760619751   0.760591855   0.760566170
#> vs            2.577201122   2.577154406   2.577116847   2.577087510
#> wt           -2.426009876  -2.425916483  -2.425827524  -2.425742964
#>                    0.0142        0.0132        0.0123        0.0115
#> (Intercept) -17.092557667 -17.092067042 -17.091680806 -17.091388048
#> am            4.242868037   4.242802899   4.242742810   4.242687485
#> carb         -1.686302931  -1.686297496  -1.686294927  -1.686294877
#> cyl           1.651764800   1.651755634   1.651751948   1.651753071
#> disp          0.006119432   0.006118457   0.006117547   0.006116698
#> drat          3.625425073   3.625444978   3.625467786   3.625492997
#> gear          2.463069779   2.463036700   2.463011725   2.462993978
#> hp            0.000000000   0.000000000   0.000000000   0.000000000
#> qsec          0.760542568   0.760520925   0.760501120   0.760483039
#> vs            2.577065528   2.577050100   2.577040489   2.577036016
#> wt           -2.425662746  -2.425586793  -2.425515014  -2.425447306
#>                    0.0107        0.0100        0.0093        0.0087
#> (Intercept) -17.091178683 -17.091043411 -17.090973674 -17.090961607
#> am            4.242636647   4.242590030   4.242547373   4.242508427
#> carb         -1.686297028  -1.686301088  -1.686306793  -1.686313900
#> cyl           1.651758390   1.651767342   1.651779417   1.651794150
#> disp          0.006115908   0.006115174   0.006114493   0.006113862
#> drat          3.625520160   3.625548870   3.625578764   3.625609516
#> gear          2.462982647   2.462976989   2.462976319   2.462980011
#> hp            0.000000000   0.000000000   0.000000000   0.000000000
#> qsec          0.760466571   0.760451608   0.760438049   0.760425796
#> vs            2.577036058   2.577040045   2.577047455   2.577057811
#> wt           -2.425383554  -2.425323632  -2.425267411  -2.425214754
#>                    0.0081        0.0076        0.0071        0.0066
#> (Intercept) -17.090999999 -17.091082257 -17.091202358 -17.091354821
#> am            4.242472954   4.242440722   4.242411512   4.242385114
#> carb         -1.686322190  -1.686331465  -1.686341545  -1.686352270
#> cyl           1.651811117   1.651829937   1.651850266   1.651871793
#> disp          0.006113279   0.006112742   0.006112247   0.006111792
#> drat          3.625640839   3.625672476   3.625704202   3.625735820
#> gear          2.462987491   2.462998236   2.463011771   2.463027664
#> hp            0.000000000   0.000000000   0.000000000   0.000000000
#> qsec          0.760414757   0.760404843   0.760395971   0.760388062
#> vs            2.577070681   2.577085672   2.577102427   2.577120625
#> wt           -2.425165522  -2.425119572  -2.425076760  -2.425036943
#>                    0.0061        0.0057        0.0053
#> (Intercept) -17.091534664 -17.033820858 -17.032713391
#> am            4.242361328   4.238597352   4.238770098
#> carb         -1.686363495  -1.658613499  -1.658398716
#> cyl           1.651894241   1.660283730   1.660135858
#> disp          0.006111375   0.006555432   0.006560259
#> drat          3.625767157   3.627681512   3.627522043
#> gear          2.463045521   2.460472657   2.460226293
#> hp            0.000000000  -0.001132631  -0.001135261
#> qsec          0.760381040   0.758203984   0.758286023
#> vs            2.577139977   2.601957825   2.601872139
#> wt           -2.424999976  -2.452369346  -2.452850119
#> 
#> $iter
#>   [1]   0   5   3   3   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   4   3   3   3   3   3   3   3   3  20  36   1   1   1 181  15
#>  [37]   1   1   1   1   1   1   1  35   1   1   1   1   1  38   1   1  81   1
#>  [55]   1   1   1   1   1   1   1   1   1   1   1   1   1  80   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1  59   1
#> 
#> $lambda
#>   [1] 5.338589466 4.978786403 4.643232863 4.330294509 4.038447152 3.766269330
#>   [7] 3.512435382 3.275708993 3.054937170 2.849044630 2.657028557 2.477953725
#>  [13] 2.310947937 2.155197781 2.009944664 1.874481120 1.748147367 1.630328085
#>  [19] 1.520449429 1.417976226 1.322409374 1.233283408 1.150164234 1.072647014
#>  [25] 1.000354195 0.932933671 0.870057064 0.811418130 0.756731263 0.705730109
#>  [31] 0.658166262 0.613808058 0.572439450 0.533858947 0.497878640 0.464323286
#>  [37] 0.433029451 0.403844715 0.376626933 0.351243538 0.327570899 0.305493717
#>  [43] 0.284904463 0.265702856 0.247795373 0.231094794 0.215519778 0.200994466
#>  [49] 0.187448112 0.174814737 0.163032808 0.152044943 0.141797623 0.132240937
#>  [55] 0.123328341 0.115016423 0.107264701 0.100035420 0.093293367 0.087005706
#>  [61] 0.081141813 0.075673126 0.070573011 0.065816626 0.061380806 0.057243945
#>  [67] 0.053385895 0.049787864 0.046432329 0.043302945 0.040384472 0.037662693
#>  [73] 0.035124354 0.032757090 0.030549372 0.028490446 0.026570286 0.024779537
#>  [79] 0.023109479 0.021551978 0.020099447 0.018744811 0.017481474 0.016303281
#>  [85] 0.015204494 0.014179762 0.013224094 0.012332834 0.011501642 0.010726470
#>  [91] 0.010003542 0.009329337 0.008700571 0.008114181 0.007567313 0.007057301
#>  [97] 0.006581663 0.006138081 0.005724394 0.005338589
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 29
#> 
#> $loss
#>   [1] 791.52571 676.62964 580.50200 500.44185 434.11736 379.51650 334.90437
#>   [8] 298.78601 269.87430 247.06190 229.39712 216.06279 206.35807 199.68265
#>  [15] 195.52298 193.44042 193.01443 193.01443 193.01443 193.01443 190.86214
#>  [22] 183.60173 177.54443 172.51649 168.36765 164.96828 162.20667 159.98662
#>  [29] 158.22533 152.94250 116.75504 116.75503 116.75503 116.75503 101.07095
#>  [36] 100.73414 100.73413 100.73413 100.73413 100.73412 100.73412 100.73412
#>  [43] 100.73412  95.19207  95.19207  95.19207  95.19207  95.19207  95.19206
#>  [50]  92.58742  92.58740  92.58740  78.53267  78.53266  78.53265  78.53264
#>  [57]  78.53264  78.53263  78.53263  78.53262  78.53262  78.53262  78.53262
#>  [64]  78.53261  78.53261  78.53261  78.53261  77.72776  77.72774  77.72773
#>  [71]  77.72772  77.72772  77.72771  77.72770  77.72770  77.72770  77.72769
#>  [78]  77.72769  77.72769  77.72769  77.72769  77.72768  77.72768  77.72768
#>  [85]  77.72768  77.72768  77.72768  77.72768  77.72768  77.72768  77.72768
#>  [92]  77.72768  77.72768  77.72768  77.72768  77.72768  77.72768  77.72768
#>  [99]  77.71775  77.71773
#> 
#> $linear.predictors
#>      5.3386   4.9788   4.6432   4.3303   4.0384   3.7663   3.5124   3.2757
#> 1  20.71429 20.86944 21.01413 21.14907 21.27492 21.39228 21.50174 21.60382
#> 2  20.71429 20.68049 20.64898 20.61959 20.59218 20.56661 20.54277 20.52054
#> 3  20.71429 20.55546 20.40733 20.26919 20.14036 20.02021 19.90816 19.80367
#> 4  20.71429 20.54434 20.38585 20.23805 20.10020 19.97165 19.85175 19.73994
#> 5  20.71429 20.48321 20.26772 20.06674 19.87931 19.70452 19.54150 19.38947
#> 6  20.71429 20.71661 20.71879 20.72081 20.72270 20.72446 20.72610 20.72764
#> 7  20.71429 20.55546 20.40733 20.26919 20.14036 20.02021 19.90816 19.80367
#> 8  20.71429 20.20536 19.73073 19.28809 18.87528 18.49030 18.13126 17.79642
#> 9  20.71429 20.36651 20.04218 19.73971 19.45762 19.19455 18.94920 18.72039
#> 10 20.71429 19.54962 18.46344 17.45047 16.50577 15.62474 14.80309 14.03681
#> 11 20.71429 19.49682 18.36141 17.30253 16.31500 15.39404 14.53514 13.73413
#> 12 20.71429 21.24454 21.73906 22.20025 22.63036 23.03148 23.40557 23.75444
#> 13 20.71429 21.44738 22.13106 22.76867 23.36330 23.91786 24.43504 24.91737
#> 14 20.71429 21.09728 21.45446 21.78757 22.09822 22.38794 22.65814 22.91012
#> 15 20.71429 20.55824 20.41270 20.27698 20.15040 20.03236 19.92227 19.81960
#> 16 20.71429 21.39181 22.02367 22.61294 23.16250 23.67502 24.15299 24.59876
#> 17 20.71429 21.27789 21.80350 22.29369 22.75084 23.17719 23.57479 23.94561
#> 18 20.71429 21.62632 22.47688 23.27012 24.00990 24.69982 25.34324 25.94329
#> 19 20.71429 20.92779 21.12690 21.31259 21.48576 21.64727 21.79789 21.93836
#> 20 20.71429 20.48321 20.26772 20.06674 19.87931 19.70452 19.54150 19.38947
#> 21 20.71429 20.92223 21.11616 21.29701 21.46568 21.62298 21.76968 21.90650
#>      3.0549   2.8490   2.6570   2.4780   2.3109    2.1552    2.0099    1.8745
#> 1  21.69902 21.78780 21.87060 21.94782 22.01983 22.086990 22.149624 22.208037
#> 2  20.49981 20.48047 20.46243 20.44561 20.42993 20.415302 20.401659 20.388937
#> 3  19.70621 19.61532 19.53056 19.45151 19.37779 19.309037 19.244918 19.185120
#> 4  19.63567 19.53842 19.44773 19.36315 19.28427 19.210703 19.142097 19.078114
#> 5  19.24769 19.11546 18.99215 18.87714 18.76989 18.669862 18.576578 18.489582
#> 6  20.72907 20.73040 20.73164 20.73280 20.73388 20.734889 20.735829 20.736706
#> 7  19.70621 19.61532 19.53056 19.45151 19.37779 19.309037 19.244918 19.185120
#> 8  17.48414 17.19292 16.92132 16.66802 16.43180 16.211497 16.006042 15.814434
#> 9  18.50700 18.30799 18.12240 17.94931 17.78789 17.637349 17.496953 17.366020
#> 10 13.32218 12.65571 12.03416 11.45450 10.91391 10.409755  9.939575  9.501084
#> 11 12.98711 12.29043 11.64071 11.03477 10.46968  9.942665  9.451173  8.992806
#> 12 24.07980 24.38323 24.66621 24.93012 25.17625 25.405783 25.619849 25.819487
#> 13 25.36719 25.78669 26.17792 26.54278 26.88305 27.200390 27.496341 27.772345
#> 14 23.14512 23.36428 23.56868 23.75929 23.93706 24.102850 24.257465 24.401659
#> 15 19.72385 19.63455 19.55127 19.47360 19.40117 19.333621 19.270623 19.211872
#> 16 25.01448 25.40218 25.76375 26.10096 26.41543 26.708717 26.982233 27.237316
#> 17 24.29143 24.61394 24.91471 25.19522 25.45682 25.700787 25.928313 26.140505
#> 18 26.50291 27.02481 27.51153 27.96545 28.38878 28.783577 29.151766 29.495141
#> 19 22.06936 22.19153 22.30547 22.41173 22.51083 22.603247 22.689437 22.769818
#> 20 19.24769 19.11546 18.99215 18.87714 18.76989 18.669862 18.576578 18.489582
#> 21 22.03409 22.15308 22.26405 22.36755 22.46407 22.554080 22.638026 22.716316
#>      1.7481   1.6303   1.5204   1.4180    1.3224    1.2333    1.1502    1.0726
#> 1  22.24898 22.24898 22.24898 22.24898 22.225115 22.138994 22.058673 21.983766
#> 2  20.38002 20.38002 20.38002 20.38002 20.409314 20.515022 20.613611 20.705556
#> 3  19.14321 19.14321 19.14321 19.14321 19.122990 19.050043 18.982008 18.918559
#> 4  19.03327 19.03327 19.03327 19.03327 19.080571 19.251266 19.410465 19.558935
#> 5  18.42860 18.42860 18.42860 18.42860 18.384376 18.224789 18.075950 17.937141
#> 6  20.73732 20.73732 20.73732 20.73732 20.839062 21.206189 21.548590 21.867915
#> 7  19.14321 19.14321 19.14321 19.14321 19.149946 19.174269 19.196954 19.218111
#> 8  15.68013 15.68013 15.68013 15.68013 15.671986 15.642605 15.615203 15.589647
#> 9  17.27424 17.27424 17.27424 17.27424 17.276863 17.286316 17.295132 17.303354
#> 10  9.19373  9.19373  9.19373  9.19373  9.205427  9.247634  9.286999  9.323711
#> 11  8.67152  8.67152  8.67152  8.67152  8.672037  8.673904  8.675645  8.677268
#> 12 25.95942 25.95942 25.95942 25.95942 25.982789 26.067112 26.145755 26.219099
#> 13 27.96581 27.96581 27.96581 27.96581 27.995872 28.104361 28.205544 28.299907
#> 14 24.50273 24.50273 24.50273 24.50273 24.539182 24.670720 24.793398 24.907808
#> 15 19.17069 19.17069 19.17069 19.17069 19.156339 19.104553 19.056255 19.011212
#> 16 27.41611 27.41611 27.41611 27.41611 27.425764 27.460592 27.493075 27.523368
#> 17 26.28924 26.28924 26.28924 26.28924 26.253883 26.126307 26.007324 25.896360
#> 18 29.73582 29.73582 29.73582 29.73582 29.700630 29.573634 29.455191 29.344730
#> 19 22.82616 22.82616 22.82616 22.82616 22.769605 22.565527 22.375194 22.197689
#> 20 18.42860 18.42860 18.42860 18.42860 18.358262 18.104444 17.867720 17.646951
#> 21 22.77119 22.77119 22.77119 22.77119 22.779985 22.811718 22.841313 22.868913
#>       1.0004    0.9329    0.8701    0.8114    0.7567    0.7057   0.6582
#> 1  21.913907 21.848757 21.787997 21.731332 21.678487 21.699892 22.90256
#> 2  20.791304 20.871272 20.945851 21.015404 21.080269 21.001770 20.32032
#> 3  18.859386 18.804202 18.752736 18.704739 18.659977 18.530007 16.60826
#> 4  19.697398 19.826529 19.946957 20.059269 20.164011 20.052734 20.34394
#> 5  17.807689 17.686961 17.574369 17.469366 17.371440 17.226132 14.72330
#> 6  22.165718 22.443451 22.702465 22.944022 23.169299 23.464253 24.67950
#> 7  19.237841 19.256241 19.273402 19.289405 19.304331 19.447437 18.13208
#> 8  15.565814 15.543587 15.522858 15.503527 15.485498 15.419681 14.73361
#> 9  17.311022 17.318173 17.324842 17.331062 17.336862 17.262071 16.51907
#> 10  9.357948  9.389878  9.419656  9.447427  9.473326  9.525205 11.06550
#> 11  8.678782  8.680194  8.681511  8.682739  8.683885  8.810434 10.04793
#> 12 26.287499 26.351289 26.410780 26.466261 26.518004 26.612682 28.31251
#> 13 28.387910 28.469982 28.546523 28.617906 28.684477 28.795080 30.17262
#> 14 25.014508 25.114016 25.206818 25.293365 25.374079 25.411344 23.76919
#> 15 18.969204 18.930028 18.893492 18.859419 18.827642 18.708359 16.96007
#> 16 27.551620 27.577967 27.602539 27.625455 27.646826 27.688344 28.61277
#> 17 25.792874 25.696363 25.606356 25.522416 25.444132 25.521328 25.23648
#> 18 29.241714 29.145642 29.056044 28.972485 28.894557 28.718281 27.79054
#> 19 22.032147 21.877762 21.733782 21.599506 21.474279 21.342003 21.48086
#> 20 17.441061 17.249047 17.069974 16.902971 16.747222 16.647038 17.45445
#> 21 22.894654 22.918659 22.941047 22.961926 22.981398 23.115926 25.13443
#>      0.6138   0.5724   0.5339   0.4979   0.4643   0.4330   0.4038   0.3766
#> 1  22.90274 22.90285 22.90292 22.69443 22.71527 22.71514 22.71504 22.71497
#> 2  20.32020 20.32012 20.32007 20.79049 20.77534 20.77537 20.77540 20.77542
#> 3  16.60802 16.60787 16.60777 17.63299 17.44333 17.44341 17.44347 17.44351
#> 4  20.34392 20.34390 20.34389 20.64298 20.70615 20.70610 20.70607 20.70605
#> 5  14.72301 14.72283 14.72271 14.66168 14.49460 14.49492 14.49515 14.49532
#> 6  24.67959 24.67964 24.67968 22.38938 22.79611 22.79635 22.79652 22.79665
#> 7  18.13193 18.13183 18.13177 16.67541 16.75787 16.75826 16.75854 16.75875
#> 8  14.73352 14.73346 14.73343 15.24901 15.16731 15.16732 15.16732 15.16733
#> 9  16.51897 16.51890 16.51886 16.44948 16.42937 16.42948 16.42957 16.42963
#> 10 11.06570 11.06582 11.06589 11.48933 11.53431 11.53405 11.53386 11.53372
#> 11 10.04810 10.04821 10.04828 10.85130 10.83930 10.83900 10.83878 10.83862
#> 12 28.31271 28.31284 28.31292 28.92753 28.98470 28.98439 28.98416 28.98400
#> 13 30.17278 30.17288 30.17295 30.21948 30.32122 30.32104 30.32092 30.32082
#> 14 23.76894 23.76879 23.76870 23.23399 23.27867 23.27898 23.27921 23.27937
#> 15 16.95984 16.95970 16.95961 17.83529 17.67401 17.67409 17.67415 17.67420
#> 16 28.61288 28.61295 28.61299 29.27139 29.27157 29.27134 29.27118 29.27106
#> 17 25.23649 25.23650 25.23650 26.31698 26.16499 26.16482 26.16469 26.16460
#> 18 27.79043 27.79037 27.79033 28.17353 28.04354 28.04360 28.04364 28.04367
#> 19 21.48092 21.48095 21.48098 20.13949 20.14913 20.14932 20.14945 20.14955
#> 20 17.45461 17.45471 17.45478 15.51681 15.57553 15.57572 15.57585 15.57594
#> 21 25.13470 25.13486 25.13497 25.83902 25.87769 25.87731 25.87703 25.87683
#>      0.3512   0.3276   0.3055   0.2849   0.2657   0.2478   0.2311   0.2155
#> 1  22.71491 22.71487 22.71484 22.71482 22.72959 22.72943 22.72931 22.72923
#> 2  20.77543 20.77545 20.77546 20.77546 20.21835 20.21843 20.21849 20.21854
#> 3  17.44355 17.44357 17.44359 17.44361 17.46470 17.46486 17.46497 17.46504
#> 4  20.70603 20.70602 20.70601 20.70601 19.45893 19.45891 19.45891 19.45893
#> 5  14.49545 14.49554 14.49560 14.49565 14.68895 14.68926 14.68948 14.68963
#> 6  22.79674 22.79680 22.79685 22.79688 23.07985 23.07995 23.08002 23.08008
#> 7  16.75890 16.75901 16.75908 16.75914 17.81929 17.81959 17.81979 17.81992
#> 8  15.16733 15.16733 15.16734 15.16734 15.07560 15.07565 15.07568 15.07570
#> 9  16.42968 16.42971 16.42973 16.42975 16.11611 16.11622 16.11630 16.11637
#> 10 11.53362 11.53354 11.53349 11.53345 11.37214 11.37191 11.37174 11.37163
#> 11 10.83850 10.83841 10.83835 10.83831 11.40849 11.40826 11.40808 11.40794
#> 12 28.98387 28.98379 28.98372 28.98367 28.96304 28.96278 28.96259 28.96246
#> 13 30.32075 30.32070 30.32066 30.32063 30.35838 30.35821 30.35809 30.35800
#> 14 23.27950 23.27959 23.27965 23.27970 23.55502 23.55532 23.55552 23.55566
#> 15 17.67424 17.67426 17.67428 17.67430 17.63917 17.63933 17.63943 17.63950
#> 16 29.27098 29.27091 29.27087 29.27083 29.26471 29.26454 29.26441 29.26432
#> 17 26.16454 26.16449 26.16446 26.16444 27.23547 27.23540 27.23532 27.23524
#> 18 28.04369 28.04371 28.04372 28.04373 27.55419 27.55426 27.55431 27.55436
#> 19 20.14961 20.14966 20.14970 20.14973 19.65081 19.65085 19.65090 19.65095
#> 20 15.57601 15.57606 15.57609 15.57612 15.16385 15.16384 15.16385 15.16388
#> 21 25.87668 25.87657 25.87649 25.87643 26.18335 26.18303 26.18279 26.18261
#>      0.2010   0.1874   0.1748   0.1630   0.1520   0.1418   0.1322   0.1233
#> 1  22.72917 22.72914 22.23383 22.23353 22.23328 21.37906 21.37888 21.37871
#> 2  20.21858 20.21861 20.59334 20.59342 20.59349 20.70444 20.70454 20.70462
#> 3  17.46509 17.46512 17.22095 17.22111 17.22124 18.09452 18.09471 18.09489
#> 4  19.45896 19.45898 19.65026 19.64998 19.64975 19.64130 19.64139 19.64148
#> 5  14.68973 14.68981 14.74526 14.74553 14.74574 13.70459 13.70439 13.70421
#> 6  23.08012 23.08015 22.57503 22.57459 22.57423 23.24415 23.24406 23.24398
#> 7  17.82000 17.82005 18.60488 18.60546 18.60592 19.59888 19.59893 19.59898
#> 8  15.07571 15.07572 14.89129 14.89130 14.89130 15.07880 15.07886 15.07891
#> 9  16.11641 16.11645 15.66562 15.66540 15.66523 16.02488 16.02492 16.02496
#> 10 11.37155 11.37149 11.37252 11.37239 11.37229 10.93884 10.93881 10.93879
#> 11 11.40783 11.40775 11.64615 11.64640 11.64660 11.42942 11.42943 11.42944
#> 12 28.96236 28.96229 29.04423 29.04418 29.04415 29.30752 29.30762 29.30771
#> 13 30.35794 30.35790 30.21756 30.21742 30.21730 30.76978 30.76986 30.76994
#> 14 23.55576 23.55582 23.61551 23.61569 23.61583 21.15310 21.15262 21.15219
#> 15 17.63955 17.63958 17.31633 17.31641 17.31648 18.28776 18.28796 18.28814
#> 16 29.26426 29.26421 29.41162 29.41167 29.41171 29.43832 29.43838 29.43843
#> 17 27.23517 27.23511 26.74230 26.74255 26.74276 27.00509 27.00512 27.00515
#> 18 27.55440 27.55443 28.01578 28.01592 28.01604 28.59409 28.59426 28.59443
#> 19 19.65100 19.65104 19.41138 19.41099 19.41068 18.56189 18.56168 18.56148
#> 20 15.16391 15.16395 15.35593 15.35562 15.35538 16.09573 16.09578 16.09582
#> 21 26.18248 26.18239 26.67024 26.67044 26.67059 25.94785 25.94779 25.94775
#>      0.1150   0.1073   0.1000   0.0933   0.0870   0.0811   0.0757   0.0706
#> 1  21.37856 21.37842 21.37830 21.37818 21.37808 21.37798 21.37789 21.37782
#> 2  20.70470 20.70478 20.70484 20.70490 20.70496 20.70501 20.70505 20.70510
#> 3  18.09505 18.09519 18.09532 18.09544 18.09555 18.09565 18.09574 18.09582
#> 4  19.64156 19.64163 19.64169 19.64175 19.64181 19.64186 19.64190 19.64194
#> 5  13.70405 13.70390 13.70376 13.70363 13.70352 13.70342 13.70332 13.70323
#> 6  23.24391 23.24384 23.24378 23.24372 23.24367 23.24362 23.24358 23.24354
#> 7  19.59903 19.59907 19.59911 19.59914 19.59917 19.59920 19.59923 19.59925
#> 8  15.07895 15.07900 15.07904 15.07907 15.07910 15.07913 15.07916 15.07918
#> 9  16.02499 16.02502 16.02505 16.02507 16.02509 16.02512 16.02513 16.02515
#> 10 10.93877 10.93874 10.93872 10.93871 10.93869 10.93868 10.93866 10.93865
#> 11 11.42944 11.42945 11.42946 11.42946 11.42946 11.42947 11.42947 11.42948
#> 12 29.30779 29.30786 29.30793 29.30799 29.30805 29.30810 29.30815 29.30819
#> 13 30.77002 30.77009 30.77015 30.77020 30.77025 30.77030 30.77034 30.77038
#> 14 21.15179 21.15143 21.15109 21.15079 21.15051 21.15026 21.15003 21.14982
#> 15 18.28830 18.28845 18.28858 18.28870 18.28882 18.28892 18.28901 18.28910
#> 16 29.43848 29.43853 29.43857 29.43861 29.43864 29.43868 29.43871 29.43873
#> 17 27.00517 27.00520 27.00522 27.00524 27.00526 27.00528 27.00530 27.00531
#> 18 28.59458 28.59472 28.59484 28.59495 28.59506 28.59515 28.59524 28.59532
#> 19 18.56131 18.56115 18.56101 18.56087 18.56075 18.56064 18.56054 18.56045
#> 20 16.09586 16.09589 16.09593 16.09595 16.09598 16.09601 16.09603 16.09605
#> 21 25.94770 25.94766 25.94763 25.94759 25.94756 25.94754 25.94751 25.94749
#>      0.0658   0.0614   0.0572   0.0534   0.0498   0.0464   0.0433   0.0404
#> 1  21.37774 21.37768 21.37762 21.37756 21.25484 21.25506 21.25527 21.25547
#> 2  20.70514 20.70517 20.70520 20.70523 20.83123 20.83107 20.83093 20.83079
#> 3  18.09590 18.09597 18.09603 18.09609 18.36468 18.36435 18.36404 18.36375
#> 4  19.64198 19.64202 19.64205 19.64208 19.46688 19.46684 19.46680 19.46677
#> 5  13.70315 13.70308 13.70302 13.70296 13.99319 13.99328 13.99336 13.99342
#> 6  23.24351 23.24347 23.24344 23.24342 23.41986 23.41989 23.41992 23.41995
#> 7  19.59927 19.59929 19.59931 19.59933 19.31134 19.31140 19.31148 19.31157
#> 8  15.07921 15.07923 15.07925 15.07926 14.62906 14.62917 14.62927 14.62938
#> 9  16.02517 16.02518 16.02519 16.02521 15.78994 15.78996 15.78998 15.79000
#> 10 10.93864 10.93863 10.93862 10.93861 11.21741 11.21735 11.21729 11.21723
#> 11 11.42948 11.42948 11.42949 11.42949 11.45155 11.45157 11.45158 11.45159
#> 12 29.30823 29.30826 29.30830 29.30832 29.24567 29.24559 29.24552 29.24545
#> 13 30.77041 30.77044 30.77047 30.77050 30.92016 30.91999 30.91983 30.91968
#> 14 21.14963 21.14946 21.14930 21.14915 21.17820 21.17873 21.17921 21.17965
#> 15 18.28917 18.28924 18.28931 18.28937 18.24560 18.24538 18.24518 18.24500
#> 16 29.43876 29.43878 29.43880 29.43882 29.45687 29.45679 29.45672 29.45666
#> 17 27.00532 27.00534 27.00535 27.00536 27.00692 27.00693 27.00693 27.00694
#> 18 28.59539 28.59546 28.59552 28.59557 28.71242 28.71220 28.71198 28.71179
#> 19 18.56037 18.56029 18.56022 18.56016 18.33628 18.33660 18.33689 18.33715
#> 20 16.09606 16.09608 16.09610 16.09611 16.31128 16.31112 16.31097 16.31083
#> 21 25.94747 25.94745 25.94743 25.94742 25.85662 25.85674 25.85685 25.85695
#>      0.0377   0.0351   0.0328   0.0305   0.0285   0.0266   0.0248   0.0231
#> 1  21.25564 21.25580 21.25595 21.25608 21.25620 21.25631 21.25641 21.25649
#> 2  20.83066 20.83054 20.83043 20.83032 20.83023 20.83014 20.83005 20.82997
#> 3  18.36349 18.36324 18.36302 18.36281 18.36261 18.36244 18.36227 18.36212
#> 4  19.46673 19.46670 19.46667 19.46664 19.46662 19.46659 19.46657 19.46655
#> 5  13.99347 13.99352 13.99355 13.99358 13.99360 13.99361 13.99362 13.99362
#> 6  23.41997 23.41999 23.42001 23.42003 23.42004 23.42005 23.42007 23.42008
#> 7  19.31166 19.31176 19.31186 19.31196 19.31207 19.31218 19.31229 19.31239
#> 8  14.62948 14.62958 14.62968 14.62978 14.62987 14.62997 14.63005 14.63014
#> 9  15.79003 15.79005 15.79008 15.79011 15.79014 15.79017 15.79020 15.79023
#> 10 11.21716 11.21709 11.21702 11.21695 11.21688 11.21681 11.21674 11.21668
#> 11 11.45160 11.45162 11.45163 11.45164 11.45165 11.45166 11.45167 11.45167
#> 12 29.24539 29.24534 29.24529 29.24525 29.24521 29.24518 29.24516 29.24513
#> 13 30.91954 30.91941 30.91930 30.91919 30.91909 30.91900 30.91892 30.91885
#> 14 21.18004 21.18040 21.18073 21.18102 21.18128 21.18151 21.18172 21.18191
#> 15 18.24483 18.24469 18.24456 18.24444 18.24433 18.24424 18.24416 18.24408
#> 16 29.45660 29.45655 29.45650 29.45645 29.45641 29.45638 29.45634 29.45631
#> 17 27.00695 27.00696 27.00696 27.00697 27.00698 27.00699 27.00700 27.00700
#> 18 28.71161 28.71144 28.71128 28.71113 28.71100 28.71087 28.71076 28.71065
#> 19 18.33739 18.33761 18.33782 18.33800 18.33816 18.33831 18.33845 18.33857
#> 20 16.31069 16.31057 16.31046 16.31035 16.31025 16.31016 16.31008 16.31000
#> 21 25.85705 25.85714 25.85722 25.85729 25.85736 25.85743 25.85749 25.85754
#>      0.0216   0.0201   0.0187   0.0175   0.0163   0.0152   0.0142   0.0132
#> 1  21.25657 21.25664 21.25671 21.25676 21.25681 21.25685 21.25689 21.25692
#> 2  20.82990 20.82983 20.82977 20.82971 20.82966 20.82961 20.82957 20.82953
#> 3  18.36199 18.36187 18.36175 18.36165 18.36156 18.36148 18.36140 18.36134
#> 4  19.46653 19.46651 19.46649 19.46647 19.46645 19.46644 19.46643 19.46641
#> 5  13.99362 13.99361 13.99361 13.99359 13.99358 13.99356 13.99355 13.99353
#> 6  23.42008 23.42009 23.42010 23.42010 23.42010 23.42011 23.42011 23.42011
#> 7  19.31250 19.31260 19.31271 19.31281 19.31291 19.31300 19.31310 19.31319
#> 8  14.63022 14.63030 14.63038 14.63045 14.63052 14.63059 14.63065 14.63071
#> 9  15.79025 15.79028 15.79031 15.79034 15.79037 15.79039 15.79042 15.79044
#> 10 11.21661 11.21654 11.21648 11.21642 11.21636 11.21631 11.21625 11.21620
#> 11 11.45168 11.45169 11.45169 11.45170 11.45171 11.45171 11.45172 11.45172
#> 12 29.24511 29.24510 29.24508 29.24507 29.24506 29.24506 29.24505 29.24505
#> 13 30.91878 30.91872 30.91867 30.91862 30.91858 30.91854 30.91850 30.91847
#> 14 21.18207 21.18222 21.18234 21.18245 21.18254 21.18262 21.18269 21.18274
#> 15 18.24402 18.24397 18.24392 18.24388 18.24385 18.24382 18.24380 18.24378
#> 16 29.45629 29.45626 29.45624 29.45622 29.45621 29.45619 29.45618 29.45617
#> 17 27.00701 27.00702 27.00703 27.00704 27.00704 27.00705 27.00706 27.00706
#> 18 28.71056 28.71047 28.71039 28.71031 28.71025 28.71019 28.71013 28.71008
#> 19 18.33868 18.33877 18.33886 18.33894 18.33900 18.33906 18.33911 18.33915
#> 20 16.30993 16.30987 16.30981 16.30975 16.30970 16.30965 16.30961 16.30958
#> 21 25.85759 25.85764 25.85768 25.85771 25.85775 25.85778 25.85781 25.85783
#>      0.0123   0.0115   0.0107   0.0100   0.0093   0.0087   0.0081   0.0076
#> 1  21.25695 21.25697 21.25699 21.25700 21.25702 21.25703 21.25703 21.25704
#> 2  20.82949 20.82946 20.82943 20.82940 20.82938 20.82935 20.82933 20.82932
#> 3  18.36128 18.36123 18.36118 18.36114 18.36110 18.36107 18.36105 18.36102
#> 4  19.46640 19.46639 19.46638 19.46637 19.46636 19.46636 19.46635 19.46634
#> 5  13.99351 13.99349 13.99347 13.99344 13.99342 13.99340 13.99338 13.99336
#> 6  23.42011 23.42011 23.42011 23.42011 23.42011 23.42011 23.42011 23.42011
#> 7  19.31327 19.31335 19.31343 19.31351 19.31358 19.31365 19.31372 19.31378
#> 8  14.63076 14.63082 14.63087 14.63091 14.63096 14.63100 14.63104 14.63107
#> 9  15.79046 15.79049 15.79051 15.79053 15.79055 15.79057 15.79058 15.79060
#> 10 11.21615 11.21610 11.21606 11.21601 11.21597 11.21593 11.21590 11.21586
#> 11 11.45172 11.45173 11.45173 11.45173 11.45174 11.45174 11.45174 11.45174
#> 12 29.24505 29.24505 29.24505 29.24505 29.24505 29.24505 29.24506 29.24506
#> 13 30.91845 30.91842 30.91840 30.91839 30.91837 30.91836 30.91835 30.91834
#> 14 21.18279 21.18282 21.18285 21.18287 21.18289 21.18289 21.18290 21.18290
#> 15 18.24376 18.24375 18.24375 18.24374 18.24374 18.24374 18.24374 18.24375
#> 16 29.45616 29.45615 29.45614 29.45614 29.45613 29.45613 29.45612 29.45612
#> 17 27.00707 27.00707 27.00708 27.00708 27.00709 27.00709 27.00710 27.00710
#> 18 28.71004 28.71000 28.70996 28.70993 28.70990 28.70988 28.70985 28.70984
#> 19 18.33919 18.33922 18.33925 18.33927 18.33929 18.33930 18.33931 18.33932
#> 20 16.30954 16.30951 16.30948 16.30946 16.30944 16.30942 16.30940 16.30939
#> 21 25.85785 25.85787 25.85789 25.85790 25.85792 25.85793 25.85794 25.85795
#>      0.0071   0.0066   0.0061   0.0057   0.0053
#> 1  21.25704 21.25704 21.25704 21.30022 21.30036
#> 2  20.82930 20.82929 20.82927 20.84773 20.84786
#> 3  18.36100 18.36099 18.36098 18.37006 18.37020
#> 4  19.46634 19.46633 19.46633 19.46678 19.46675
#> 5  13.99334 13.99332 13.99330 13.97772 13.97793
#> 6  23.42011 23.42010 23.42010 23.40776 23.40780
#> 7  19.31384 19.31389 19.31395 19.35615 19.35586
#> 8  14.63111 14.63114 14.63117 14.60673 14.60640
#> 9  15.79062 15.79063 15.79065 15.77284 15.77270
#> 10 11.21583 11.21580 11.21577 11.24406 11.24433
#> 11 11.45174 11.45175 11.45175 11.43671 11.43662
#> 12 29.24507 29.24507 29.24508 29.24026 29.24019
#> 13 30.91833 30.91833 30.91833 30.92059 30.92068
#> 14 21.18289 21.18288 21.18287 21.15851 21.15861
#> 15 18.24375 18.24376 18.24377 18.25583 18.25579
#> 16 29.45612 29.45612 29.45612 29.45993 29.45994
#> 17 27.00711 27.00711 27.00711 27.00116 27.00102
#> 18 28.70982 28.70980 28.70979 28.70801 28.70803
#> 19 18.33932 18.33932 18.33932 18.36079 18.36085
#> 20 16.30937 16.30936 16.30935 16.27107 16.27114
#> 21 25.85795 25.85796 25.85797 25.83706 25.83694
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.4 18.7 18.1 14.3 22.8 19.2 16.4 15.2 10.4 14.7 32.4 33.9 21.5 15.2
#> [16] 27.3 26.0 30.4 19.7 15.0 21.4
#> 
#> $X
#>               am       carb       cyl        disp        drat       gear
#>  [1,]  1.1547005  0.6889986  0.000000 -0.48390622  0.69382135  0.3735437
#>  [2,] -0.8660254 -0.9804980  0.000000  0.33893027 -1.02278837 -0.9338592
#>  [3,] -0.8660254 -0.4239992  1.145644  1.19535192 -0.87624852 -0.9338592
#>  [4,] -0.8660254 -0.9804980  0.000000  0.06185268 -1.69268485 -0.9338592
#>  [5,] -0.8660254  0.6889986  1.145644  1.19535192 -0.75064293 -0.9338592
#>  [6,] -0.8660254 -0.4239992 -1.145644 -0.64511500  0.73568988  0.3735437
#>  [7,] -0.8660254  0.6889986  0.000000 -0.42009441  0.73568988  0.3735437
#>  [8,] -0.8660254  0.1324997  1.145644  0.48838424 -1.04372264 -0.9338592
#>  [9,] -0.8660254  0.1324997  1.145644  0.48838424 -1.04372264 -0.9338592
#> [10,] -0.8660254  0.6889986  1.145644  2.13573649 -1.33680235 -0.9338592
#> [11,] -0.8660254  0.6889986  1.145644  1.86705518 -0.70877440 -0.9338592
#> [12,]  1.1547005 -0.9804980 -1.145644 -1.16652466  1.07063812  0.3735437
#> [13,]  1.1547005 -0.9804980 -1.145644 -1.23033647  1.36371783  0.3735437
#> [14,] -0.8660254 -0.9804980 -1.145644 -0.81891822  0.27513605 -0.9338592
#> [15,] -0.8660254 -0.4239992  1.145644  0.72515964 -0.87624852 -0.9338592
#> [16,]  1.1547005 -0.9804980 -1.145644 -1.16400577  1.07063812  0.3735437
#> [17,]  1.1547005 -0.4239992 -1.145644 -0.81723896  1.80333739  1.6809466
#> [18,]  1.1547005 -0.4239992 -1.145644 -1.02882549  0.42167591  1.6809466
#> [19,]  1.1547005  1.8019964  0.000000 -0.60985058  0.10766193  1.6809466
#> [20,]  1.1547005  2.9149942  1.145644  0.69997077 -0.05981219  1.6809466
#> [21,]  1.1547005 -0.4239992 -1.145644 -0.81136156  1.13344092  0.3735437
#>                hp        qsec         vs           wt
#>  [1,] -0.50718456 -0.57188715 -0.9534626 -0.287472090
#>  [2,] -0.50718456  0.74917217  1.0488088  0.062613314
#>  [3,]  0.46107687 -0.57188715 -0.9534626  0.294287478
#>  [4,] -0.58166621  1.17496815  1.0488088  0.314880737
#>  [5,]  1.50381994 -1.21604004 -0.9534626  0.428143661
#>  [6,] -0.73062950  2.63795946  1.0488088 -0.004314778
#>  [7,] -0.31353227  0.12685497  1.0488088  0.294287478
#>  [8,]  0.53555852 -0.36444809 -0.9534626  0.942975137
#>  [9,]  0.53555852 -0.03691272 -0.9534626  0.644372881
#> [10,]  0.90796676 -0.04783056 -0.9534626  2.157977420
#> [11,]  1.28037500 -0.35353024 -0.9534626  2.255795400
#> [12,] -1.16262306  0.76554894  1.0488088 -0.982494582
#> [13,] -1.17751939  1.00028262  1.0488088 -1.358321559
#> [14,] -0.70083684  1.06033077  1.0488088 -0.709633900
#> [15,]  0.08866863 -0.41903731 -0.9534626  0.289139163
#> [16,] -1.16262306  0.45439034  1.0488088 -1.255355264
#> [17,] -0.79021482 -0.74657268 -0.9534626 -1.044274359
#> [18,] -0.46249557 -0.63739423  1.0488088 -1.689873029
#> [19,]  0.46107687 -1.40164342 -0.9534626 -0.395586700
#> [20,]  2.84448961 -1.89294647 -0.9534626  0.428143661
#> [21,] -0.52208089  0.29062265  1.0488088 -0.385290070
#> attr(,"center")
#>  [1]   0.4285714   2.7619048   6.0000000 217.6333333   3.5685714   3.7142857
#>  [7] 144.0476190  18.0676190   0.4761905   3.1541905
#> attr(,"scale")
#>  [1]   0.4948717   1.7969488   1.7457431 119.1002112   0.4776857   0.7648752
#>  [7]  67.1306304   1.8318632   0.4994328   0.9711916
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 35
#> 
#> $lambda.min
#> [1] 0.4978786
#> 
#> $null.dev
#> [1] 37.6917
#> 
#> $Bias
#> [1] 8.17257
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 7.395779