Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 40.14678 37.25205 33.10079 29.23955 26.15668 23.73223 21.86312 20.46088
#>   [9] 19.44964 18.76443 18.34969 18.15798 18.14884 18.28789 18.54592 18.82090
#>  [17] 18.38978 18.04739 17.97123 17.78398 17.58083 17.39853 17.23476 17.04809
#>  [25] 16.91263 16.73684 16.59651 14.31227 13.22210 12.51128 12.23480 11.81621
#>  [33] 11.55021 11.60183 11.67974 11.38005 11.26099 10.85399 10.89922 10.94337
#>  [41] 10.97308 11.89717 11.82253 13.42111 13.20958 13.34537 14.42001 14.19510
#>  [49] 14.79872 14.75586 17.07831 17.07896 17.08152 23.66432 28.11599 28.13122
#>  [57] 28.09425 29.46608 29.38410 29.38477 30.96502 31.14812 31.20995 31.26858
#>  [65] 31.26814 32.58384 32.79478 32.97148 33.19443 33.19316 32.87132 33.02986
#>  [73] 32.96635 33.17130 33.16984 33.16840 33.18534 33.29095 33.28966 33.28874
#>  [81] 33.28977 33.60691 33.60569 33.60427 33.29044 33.28648 33.28488 33.28330
#>  [89] 33.28176 33.25538 33.25392 33.37418 33.37293 33.43372 33.43259 33.43148
#>  [97] 33.43038 33.42931 33.42825 33.42721
#> 
#> $cvse
#>   [1] 11.638629 11.550476 10.681635  9.732605  8.866871  8.077206  7.361794
#>   [8]  6.723189  6.167361  5.702572  5.337756  5.080245  4.933137  4.893253
#>  [15]  4.950786  5.057945  4.910823  4.815487  4.781922  4.723762  4.666554
#>  [22]  4.616055  4.571237  4.538556  4.491741  4.458331  4.430709  4.337976
#>  [29]  4.323491  4.308958  4.261943  4.241940  4.227911  4.197753  4.126219
#>  [36]  4.064936  3.910300  3.505350  3.499011  3.493051  3.488501  3.597319
#>  [43]  3.504899  3.503599  3.458299  3.479368  4.002597  4.038631  4.083412
#>  [50]  4.101138  4.994191  4.994448  4.994359  8.793609 11.062835 11.060753
#>  [57] 11.057701 11.935388 11.903531 11.902734 13.019509 13.011162 13.013051
#>  [64] 13.005885 13.005398 13.720140 13.692048 13.680088 13.664222 13.663046
#>  [71] 13.660676 13.678588 13.668173 13.646013 13.644969 13.643941 13.640834
#>  [78] 13.632811 13.631818 13.630842 13.631152 13.683341 13.682416 13.681368
#>  [85] 13.495046 13.492164 13.490966 13.489792 13.488640 13.490378 13.489279
#>  [92] 13.478605 13.477533 13.474486 13.473450 13.472433 13.471435 13.470456
#>  [99] 13.469494 13.468550
#> 
#> $fold
#>  [1]  3  9 10  6  1  5  2  9  7 10  8  2  5  4  1  3  7  8  4  1  6
#> 
#> $lambda
#>   [1] 5.504634372 5.133640439 4.787650255 4.464978651 4.164054034 3.883410729
#>   [7] 3.621681843 3.377592608 3.149954171 2.937657802 2.739669498 2.555024943
#>  [13] 2.382824814 2.222230396 2.072459505 1.932782671 1.802519587 1.681035800
#>  [19] 1.567739613 1.462079210 1.363539965 1.271641935 1.185937525 1.106009305
#>  [25] 1.031467979 0.961950490 0.897118246 0.836655479 0.780267700 0.727680268
#>  [31] 0.678637054 0.632899188 0.590243901 0.550463437 0.513364044 0.478765025
#>  [37] 0.446497865 0.416405403 0.388341073 0.362168184 0.337759261 0.314995417
#>  [43] 0.293765780 0.273966950 0.255502494 0.238282481 0.222223040 0.207245951
#>  [49] 0.193278267 0.180251959 0.168103580 0.156773961 0.146207921 0.136353997
#>  [55] 0.127164194 0.118593752 0.110600930 0.103146798 0.096195049 0.089711825
#>  [61] 0.083665548 0.078026770 0.072768027 0.067863705 0.063289919 0.059024390
#>  [67] 0.055046344 0.051336404 0.047876503 0.044649787 0.041640540 0.038834107
#>  [73] 0.036216818 0.033775926 0.031499542 0.029376578 0.027396695 0.025550249
#>  [79] 0.023828248 0.022222304 0.020724595 0.019327827 0.018025196 0.016810358
#>  [85] 0.015677396 0.014620792 0.013635400 0.012716419 0.011859375 0.011060093
#>  [91] 0.010314680 0.009619505 0.008971182 0.008366555 0.007802677 0.007276803
#>  [97] 0.006786371 0.006328992 0.005902439 0.005504634
#> 
#> $fit
#> $beta
#>               5.5046     5.1336     4.7877     4.4650   4.1641    3.8834
#> (Intercept) 20.51905 22.3749647 24.1586817 25.8221821 27.37357 28.820396
#> am           0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#> cyl          0.00000 -0.3093195 -0.6066057 -0.8838558 -1.14242 -1.383558
#> disp         0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#> wt           0.00000  0.0000000  0.0000000  0.0000000  0.00000  0.000000
#>                3.6217    3.3776    3.1500   2.9377    2.7397    2.5550
#> (Intercept) 30.169713 31.428090 32.601656 33.69613 34.716837 35.668753
#> am           0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> cyl         -1.608444 -1.818174 -2.013768 -2.19618 -2.366298 -2.524951
#> disp         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#>                2.3828    2.2222    2.0725    1.9328    1.8025    1.6810
#> (Intercept) 36.556513 37.384441 38.156570 38.876659 39.400298 39.400298
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.672911 -2.810899 -2.939587 -3.059602 -3.146875 -3.146875
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.5677    1.4621    1.3635    1.2716    1.1859    1.1060
#> (Intercept) 39.400298 39.400298 39.400298 39.400298 39.400298 39.400298
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -3.146875 -3.146875 -3.146875 -3.146875 -3.146875 -3.146875
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                 1.0315    0.9620   0.8971    0.8367    0.7803    0.7277
#> (Intercept) 39.7472501 40.402631 40.61130 40.610967 40.610760 40.610632
#> am           0.0000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> carb         0.0000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> cyl         -2.7838035 -2.094225 -1.87413 -1.873932 -1.873809 -1.873733
#> disp         0.0000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> drat         0.0000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> gear         0.0000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> hp           0.0000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> qsec         0.0000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> vs           0.0000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> wt          -0.8133025 -2.356847 -2.84934 -2.849615 -2.849786 -2.849892
#>                0.6786    0.6329    0.5902    0.5505    0.5134    0.4788
#> (Intercept) 40.610553 40.610503 40.610472 40.610453 40.610441 40.610434
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.873686 -1.873656 -1.873638 -1.873627 -1.873620 -1.873615
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -2.849958 -2.849998 -2.850024 -2.850040 -2.850049 -2.850056
#>                  0.4465      0.4164     0.3883     0.3622     0.3378     0.3150
#> (Intercept) 40.60718132 40.55416076 40.5045263 40.4585387 40.4153531 40.3750702
#> am           0.00000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb        -0.00451082 -0.07264106 -0.1362972 -0.1954139 -0.2507921 -0.3024444
#> cyl         -1.87028701 -1.81770157 -1.7685125 -1.7228950 -1.6800991 -1.6401809
#> disp         0.00000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.00000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.00000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.00000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.00000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> vs           0.00000000  0.00000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.85101219 -2.86867899 -2.8852648 -2.9005790 -2.9150127 -2.9284775
#>                 0.2938     0.2740     0.2555     0.2383     0.2222     0.2072
#> (Intercept) 40.3375020 40.3024658 40.2735980 11.8296761 11.8281678 11.8270876
#> am           0.0000000  0.0000000  0.0000000  3.7917710  3.7925426  3.7930945
#> carb        -0.3506157 -0.3955404 -0.4327387 -0.4766133 -0.4767246 -0.4768042
#> cyl         -1.6029529 -1.5682339 -1.5395710  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  1.0424442  1.0424575  1.0424671
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.9410350 -2.9527462 -2.9623253 -3.3488966 -3.3485191 -3.3482492
#>                0.1933     0.1803     0.1681     0.1568     0.1462      0.1364
#> (Intercept) 11.826314 11.8257612 11.8253654 11.8250823 11.8248799 11.79251404
#> am           3.793489  3.7937713  3.7939731  3.7941173  3.7942205  3.79059511
#> carb        -0.476861 -0.4769016 -0.4769306 -0.4769513 -0.4769661 -0.47736279
#> cyl          0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.00000000
#> disp         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.00000000
#> drat         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.01005457
#> gear         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.00000000
#> hp           0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.00000000
#> qsec         1.042474  1.0424790  1.0424825  1.0424851  1.0424869  1.04204323
#> vs           0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.00000000
#> wt          -3.348056 -3.3479182 -3.3478196 -3.3477491 -3.3476986 -3.34562268
#>                 0.1272     0.1186     0.1106     0.1031     0.0962     0.0897
#> (Intercept) 11.2665620 10.7834399  8.7381382  8.3369640  8.3383075  8.3397851
#> am           3.6850730  3.5859036  3.5641597  3.4899402  3.4890732  3.4884228
#> carb        -0.4845729 -0.4910378 -0.5044551 -0.5098412 -0.5098302 -0.5098124
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.2055983  0.3864549  0.4972244  0.6429947  0.6433748  0.6435711
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         1.0310540  1.0208502  1.1628760  1.1545432  1.1544413  1.1543624
#> vs           0.0000000  0.0000000 -0.8703416 -0.8669878 -0.8667841 -0.8666412
#> wt          -3.3192325 -3.2955792 -3.4371297 -3.4149980 -3.4151824 -3.4153603
#>                 0.0837     0.0780     0.0728     0.0679     0.0633     0.0590
#> (Intercept)  8.3411321  8.3422305  8.3430571  8.3436387  8.3440228  8.3442590
#> am           3.4879387  3.4875827  3.4873247  3.4871409  3.4870121  3.4869236
#> carb        -0.5097968 -0.5097853 -0.5097772 -0.5097712 -0.5097664 -0.5097623
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.6436674  0.6437112  0.6437285  0.6437326  0.6437305  0.6437256
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         1.1543043  1.1542643  1.1542389  1.1542245  1.1542180  1.1542168
#> vs          -0.8665581 -0.8665228 -0.8665210 -0.8665397 -0.8665691 -0.8666026
#> wt          -3.4155163 -3.4156455 -3.4157487 -3.4158290 -3.4158903 -3.4159367
#>                 0.0550     0.0513     0.0479        0.0446     0.0416
#> (Intercept)  8.3443913  8.3444544  8.3444741  8.3492703793 12.7524248
#> am           3.4868640  3.4868249  3.4868000  3.4867847835  3.1677093
#> carb        -0.5097586 -0.5097552 -0.5097521 -0.5097492543 -0.4857585
#> cyl          0.0000000  0.0000000  0.0000000 -0.0002153227 -0.3107280
#> disp         0.0000000  0.0000000  0.0000000  0.0000000000  0.0000000
#> drat         0.6437196  0.6437133  0.6437071  0.6431980034  0.4621628
#> gear         0.0000000  0.0000000  0.0000000  0.0000000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000000  0.0000000
#> qsec         1.1542188  1.1542225  1.1542269  1.1541267855  1.0275281
#> vs          -0.8666358 -0.8666663 -0.8666931 -0.8668269711 -0.9851626
#> wt          -3.4159714 -3.4159973 -3.4160164 -3.4159542332 -3.2491968
#>                 0.0388       0.0362       0.0338       0.0315      0.0294
#> (Intercept) 12.7566038 13.908045718 13.910217541 13.912307310 13.91431801
#> am           3.1670871  3.354992936  3.355001488  3.355009554  3.35501717
#> carb        -0.4857018 -0.230827873 -0.230655565 -0.230489841 -0.23033045
#> cyl         -0.3110220 -0.650486391 -0.650786705 -0.651075574 -0.65135343
#> disp         0.0000000  0.009094438  0.009099407  0.009104185  0.00910878
#> drat         0.4621289  0.441454928  0.441352767  0.441254547  0.44116011
#> gear         0.0000000 -0.431291739 -0.431559532 -0.431817098 -0.43206482
#> hp           0.0000000  0.000000000  0.000000000  0.000000000  0.00000000
#> qsec         1.0273995  1.167468354  1.167518955  1.167567574  1.16761429
#> vs          -0.9852432 -1.044441238 -1.044576487 -1.044706562 -1.04483166
#> wt          -3.2491345 -4.100217779 -4.100663311 -4.101091714 -4.10150365
#>                   0.0274       0.0256       0.0238       0.0222       0.0207
#> (Intercept) 13.916252526 13.918113655 13.921231541 13.922052396 13.922855703
#> am           3.355024378  3.355031197  3.370804075  3.370854257  3.370903341
#> carb        -0.230177166 -0.230029743 -0.048094933 -0.047934035 -0.047776613
#> cyl         -0.651620672 -0.651877711 -0.644968385 -0.645116989 -0.645262398
#> disp         0.009113198  0.009117447  0.014504914  0.014509953  0.014514883
#> drat         0.441069321  0.440982028  0.485263341  0.485217339  0.485172329
#> gear        -0.432303076 -0.432532217 -0.424666428 -0.424810782 -0.424952029
#> hp           0.000000000  0.000000000 -0.007029581 -0.007031939 -0.007034246
#> qsec         1.167659184  1.167702325  1.164167185  1.164212425  1.164256685
#> vs          -1.044951966 -1.045067666 -0.743710472 -0.743669426 -0.743629284
#> wt          -4.101899743 -4.102280610 -4.412733482 -4.413138937 -4.413535626
#>                   0.0193       0.0180       0.0168       0.0157       0.0146
#> (Intercept) 13.923641827 13.924411127 13.925163952 13.925900649 13.926621555
#> am           3.370951353  3.370998316  3.371044253  3.371089188  3.371133142
#> carb        -0.047622592 -0.047471899 -0.047324462 -0.047180212 -0.047039079
#> cyl         -0.645404680 -0.645543901 -0.645680126 -0.645813418 -0.645943841
#> disp         0.014519707  0.014524426  0.014529044  0.014533561  0.014537981
#> drat         0.485128287  0.485085195  0.485043031  0.485001775  0.484961410
#> gear        -0.425090233 -0.425225461 -0.425357775 -0.425487238 -0.425613910
#> hp          -0.007036503 -0.007038711 -0.007040871 -0.007042984 -0.007045051
#> qsec         1.164299985  1.164342347  1.164383791  1.164424337  1.164464004
#> vs          -0.743590024 -0.743551626 -0.743514072 -0.743477342 -0.743441418
#> wt          -4.413923738 -4.414303458 -4.414674968 -4.415038443 -4.415394057
#>                   0.0136       0.0127       0.0119       0.0111       0.0103
#> (Intercept) 13.927327000 13.928017311 13.928692805 13.929353795 13.930000587
#> am           3.371176137  3.371218195  3.371259335  3.371299579  3.371338947
#> carb        -0.046900997 -0.046765901 -0.046633724 -0.046504406 -0.046377884
#> cyl         -0.646071454 -0.646196319 -0.646318493 -0.646438033 -0.646554997
#> disp         0.014542305  0.014546535  0.014550674  0.014554724  0.014558686
#> drat         0.484921915  0.484883271  0.484845461  0.484808467  0.484772272
#> gear        -0.425737851 -0.425859119 -0.425977771 -0.426093864 -0.426207452
#> hp          -0.007047074 -0.007049053 -0.007050988 -0.007052882 -0.007054735
#> qsec         1.164502812  1.164540779  1.164577923  1.164614262  1.164649814
#> vs          -0.743406281 -0.743371914 -0.743338300 -0.743305422 -0.743273262
#> wt          -4.415741980 -4.416082377 -4.416415411 -4.416741241 -4.417060021
#>                   0.0096       0.0090       0.0084       0.0078       0.0073
#> (Intercept) 13.930633480 13.931252768 13.931858739 13.932451676 13.933031855
#> am           3.371377456  3.371415127  3.371451978  3.371488026  3.371523290
#> carb        -0.046254098 -0.046132988 -0.046014498 -0.045898571 -0.045785151
#> cyl         -0.646669438 -0.646781412 -0.646890970 -0.646998164 -0.647103046
#> disp         0.014562562  0.014566354  0.014570064  0.014573694  0.014577246
#> drat         0.484736857  0.484702207  0.484668306  0.484635136  0.484602682
#> gear        -0.426318589 -0.426427327 -0.426533717 -0.426637811 -0.426739657
#> hp          -0.007056547 -0.007058321 -0.007060056 -0.007061753 -0.007063413
#> qsec         1.164684596  1.164718624  1.164751915  1.164784484  1.164816348
#> vs          -0.743241807 -0.743211038 -0.743180943 -0.743151504 -0.743122708
#> wt          -4.417371906 -4.417677042 -4.417975576 -4.418267650 -4.418553405
#>                   0.0068       0.0063       0.0059       0.0055
#> (Intercept) 13.933599547 13.934155016 13.934698524 13.935230324
#> am           3.371557787  3.371591534  3.371624546  3.371656840
#> carb        -0.045674185 -0.045565619 -0.045459402 -0.045355483
#> cyl         -0.647205664 -0.647306067 -0.647404302 -0.647500416
#> disp         0.014580720  0.014584119  0.014587445  0.014590699
#> drat         0.484570930  0.484539863  0.484509468  0.484479729
#> gear        -0.426839303 -0.426936798 -0.427032186 -0.427125513
#> hp          -0.007065038 -0.007066627 -0.007068182 -0.007069703
#> qsec         1.164847522  1.164878020  1.164907857  1.164937048
#> vs          -0.743094541 -0.743066988 -0.743040037 -0.743013673
#> wt          -4.418832975 -4.419106496 -4.419374097 -4.419635907
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1   1  65  74  15   1   1   1   1   1   1   1   1   1
#>  [37]   6  13  13  12  12  12  12  12   9 152   1   1   1   1   1   1   1   6
#>  [55]  20  20  74  19   1   1   1   1   1   1   1   1   1   1   1   1  43   1
#>  [73] 168   1   1   1   1   1 178   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.504634372 5.133640439 4.787650255 4.464978651 4.164054034 3.883410729
#>   [7] 3.621681843 3.377592608 3.149954171 2.937657802 2.739669498 2.555024943
#>  [13] 2.382824814 2.222230396 2.072459505 1.932782671 1.802519587 1.681035800
#>  [19] 1.567739613 1.462079210 1.363539965 1.271641935 1.185937525 1.106009305
#>  [25] 1.031467979 0.961950490 0.897118246 0.836655479 0.780267700 0.727680268
#>  [31] 0.678637054 0.632899188 0.590243901 0.550463437 0.513364044 0.478765025
#>  [37] 0.446497865 0.416405403 0.388341073 0.362168184 0.337759261 0.314995417
#>  [43] 0.293765780 0.273966950 0.255502494 0.238282481 0.222223040 0.207245951
#>  [49] 0.193278267 0.180251959 0.168103580 0.156773961 0.146207921 0.136353997
#>  [55] 0.127164194 0.118593752 0.110600930 0.103146798 0.096195049 0.089711825
#>  [61] 0.083665548 0.078026770 0.072768027 0.067863705 0.063289919 0.059024390
#>  [67] 0.055046344 0.051336404 0.047876503 0.044649787 0.041640540 0.038834107
#>  [73] 0.036216818 0.033775926 0.031499542 0.029376578 0.027396695 0.025550249
#>  [79] 0.023828248 0.022222304 0.020724595 0.019327827 0.018025196 0.016810358
#>  [85] 0.015677396 0.014620792 0.013635400 0.012716419 0.011859375 0.011060093
#>  [91] 0.010314680 0.009619505 0.008971182 0.008366555 0.007802677 0.007276803
#>  [97] 0.006786371 0.006328992 0.005902439 0.005504634
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 24
#> 
#> $loss
#>   [1] 834.7524 716.2819 613.9617 528.7321 458.1135 399.9661 352.4450 313.9604
#>   [9] 283.1434 258.8163 239.9670 225.7263 215.3488 208.1961 203.7217 201.4592
#>  [17] 200.9718 200.9718 200.9718 200.9718 200.9718 200.9718 200.9718 200.9718
#>  [25] 170.0082 139.5847 137.6895 137.6895 137.6895 137.6895 137.6895 137.6895
#>  [33] 137.6895 137.6895 137.6895 137.6895 137.5385 135.4496 133.8239 132.5962
#>  [41] 131.6926 131.0646 130.6658 130.4564 130.4018 115.7027 115.7027 115.7027
#>  [49] 115.7027 115.7027 115.7027 115.7027 115.7027 115.6750 115.2211 114.9447
#>  [57] 113.8352 113.7901 113.7901 113.7901 113.7901 113.7901 113.7901 113.7901
#>  [65] 113.7901 113.7901 113.7901 113.7901 113.7901 113.7894 113.2748 113.2748
#>  [73] 112.2033 112.2033 112.2033 112.2033 112.2033 112.2032 111.9173 111.9173
#>  [81] 111.9172 111.9172 111.9172 111.9172 111.9172 111.9172 111.9172 111.9172
#>  [89] 111.9171 111.9171 111.9171 111.9171 111.9171 111.9171 111.9171 111.9171
#>  [97] 111.9171 111.9171 111.9171 111.9171
#> 
#> $linear.predictors
#>      5.5046   5.1336   4.7877   4.4650   4.1641   3.8834   3.6217   3.3776
#> 1  20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905
#> 2  20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905
#> 3  20.51905 21.13769 21.73226 22.28676 22.80389 23.28616 23.73594 24.15539
#> 4  20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905
#> 5  20.51905 19.90041 19.30584 18.75134 18.23421 17.75193 17.30216 16.88270
#> 6  20.51905 21.13769 21.73226 22.28676 22.80389 23.28616 23.73594 24.15539
#> 7  20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905
#> 8  20.51905 19.90041 19.30584 18.75134 18.23421 17.75193 17.30216 16.88270
#> 9  20.51905 19.90041 19.30584 18.75134 18.23421 17.75193 17.30216 16.88270
#> 10 20.51905 19.90041 19.30584 18.75134 18.23421 17.75193 17.30216 16.88270
#> 11 20.51905 21.13769 21.73226 22.28676 22.80389 23.28616 23.73594 24.15539
#> 12 20.51905 21.13769 21.73226 22.28676 22.80389 23.28616 23.73594 24.15539
#> 13 20.51905 19.90041 19.30584 18.75134 18.23421 17.75193 17.30216 16.88270
#> 14 20.51905 19.90041 19.30584 18.75134 18.23421 17.75193 17.30216 16.88270
#> 15 20.51905 21.13769 21.73226 22.28676 22.80389 23.28616 23.73594 24.15539
#> 16 20.51905 21.13769 21.73226 22.28676 22.80389 23.28616 23.73594 24.15539
#> 17 20.51905 21.13769 21.73226 22.28676 22.80389 23.28616 23.73594 24.15539
#> 18 20.51905 19.90041 19.30584 18.75134 18.23421 17.75193 17.30216 16.88270
#> 19 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905
#> 20 20.51905 19.90041 19.30584 18.75134 18.23421 17.75193 17.30216 16.88270
#> 21 20.51905 21.13769 21.73226 22.28676 22.80389 23.28616 23.73594 24.15539
#>      3.1500   2.9377   2.7397   2.5550   2.3828   2.2222   2.0725   1.9328
#> 1  20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905
#> 2  20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905
#> 3  24.54658 24.91141 25.25164 25.56895 25.86487 26.14085 26.39822 26.63825
#> 4  20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905
#> 5  16.49151 16.12669 15.78645 15.46915 15.17323 14.89725 14.63987 14.39984
#> 6  24.54658 24.91141 25.25164 25.56895 25.86487 26.14085 26.39822 26.63825
#> 7  20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905
#> 8  16.49151 16.12669 15.78645 15.46915 15.17323 14.89725 14.63987 14.39984
#> 9  16.49151 16.12669 15.78645 15.46915 15.17323 14.89725 14.63987 14.39984
#> 10 16.49151 16.12669 15.78645 15.46915 15.17323 14.89725 14.63987 14.39984
#> 11 24.54658 24.91141 25.25164 25.56895 25.86487 26.14085 26.39822 26.63825
#> 12 24.54658 24.91141 25.25164 25.56895 25.86487 26.14085 26.39822 26.63825
#> 13 16.49151 16.12669 15.78645 15.46915 15.17323 14.89725 14.63987 14.39984
#> 14 16.49151 16.12669 15.78645 15.46915 15.17323 14.89725 14.63987 14.39984
#> 15 24.54658 24.91141 25.25164 25.56895 25.86487 26.14085 26.39822 26.63825
#> 16 24.54658 24.91141 25.25164 25.56895 25.86487 26.14085 26.39822 26.63825
#> 17 24.54658 24.91141 25.25164 25.56895 25.86487 26.14085 26.39822 26.63825
#> 18 16.49151 16.12669 15.78645 15.46915 15.17323 14.89725 14.63987 14.39984
#> 19 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905
#> 20 16.49151 16.12669 15.78645 15.46915 15.17323 14.89725 14.63987 14.39984
#> 21 24.54658 24.91141 25.25164 25.56895 25.86487 26.14085 26.39822 26.63825
#>      1.8025   1.6810   1.5677   1.4621   1.3635   1.2716   1.1859   1.1060
#> 1  20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905
#> 2  20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905
#> 3  26.81280 26.81280 26.81280 26.81280 26.81280 26.81280 26.81280 26.81280
#> 4  20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905
#> 5  14.22530 14.22530 14.22530 14.22530 14.22530 14.22530 14.22530 14.22530
#> 6  26.81280 26.81280 26.81280 26.81280 26.81280 26.81280 26.81280 26.81280
#> 7  20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905
#> 8  14.22530 14.22530 14.22530 14.22530 14.22530 14.22530 14.22530 14.22530
#> 9  14.22530 14.22530 14.22530 14.22530 14.22530 14.22530 14.22530 14.22530
#> 10 14.22530 14.22530 14.22530 14.22530 14.22530 14.22530 14.22530 14.22530
#> 11 26.81280 26.81280 26.81280 26.81280 26.81280 26.81280 26.81280 26.81280
#> 12 26.81280 26.81280 26.81280 26.81280 26.81280 26.81280 26.81280 26.81280
#> 13 14.22530 14.22530 14.22530 14.22530 14.22530 14.22530 14.22530 14.22530
#> 14 14.22530 14.22530 14.22530 14.22530 14.22530 14.22530 14.22530 14.22530
#> 15 26.81280 26.81280 26.81280 26.81280 26.81280 26.81280 26.81280 26.81280
#> 16 26.81280 26.81280 26.81280 26.81280 26.81280 26.81280 26.81280 26.81280
#> 17 26.81280 26.81280 26.81280 26.81280 26.81280 26.81280 26.81280 26.81280
#> 18 14.22530 14.22530 14.22530 14.22530 14.22530 14.22530 14.22530 14.22530
#> 19 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905 20.51905
#> 20 14.22530 14.22530 14.22530 14.22530 14.22530 14.22530 14.22530 14.22530
#> 21 26.81280 26.81280 26.81280 26.81280 26.81280 26.81280 26.81280 26.81280
#>      1.0315   0.9620   0.8971   0.8367   0.7803   0.7277   0.6786   0.6329
#> 1  20.91358 21.66234 21.90125 21.90138 21.90147 21.90152 21.90155 21.90157
#> 2  20.70618 21.06135 21.17467 21.17473 21.17477 21.17479 21.17481 21.17482
#> 3  26.72517 26.55785 26.50431 26.50413 26.50402 26.50395 26.50391 26.50388
#> 4  20.23040 19.68259 19.50780 19.50771 19.50765 19.50761 19.50758 19.50757
#> 5  14.57333 15.23489 15.44612 15.44638 15.44655 15.44665 15.44672 15.44676
#> 6  26.05013 24.60166 24.13936 24.13895 24.13870 24.13854 24.13844 24.13838
#> 7  20.24667 19.72973 19.56479 19.56470 19.56464 19.56461 19.56458 19.56457
#> 8  14.44320 14.85779 14.99022 14.99045 14.99058 14.99067 14.99072 14.99076
#> 9  13.06547 10.86529 10.16344 10.16320 10.16305 10.16295 10.16290 10.16286
#> 10 13.12972 11.05149 10.38854 10.38832 10.38818 10.38810 10.38804 10.38801
#> 11 26.82277 26.84067 26.84623 26.84609 26.84599 26.84594 26.84590 26.84588
#> 12 27.11963 27.70092 27.88624 27.88619 27.88617 27.88615 27.88614 27.88613
#> 13 14.61400 15.35273 15.58858 15.58886 15.58904 15.58915 15.58921 15.58926
#> 14 14.35374 14.59854 14.67679 14.67699 14.67711 14.67718 14.67723 14.67726
#> 15 27.03830 27.46523 27.60131 27.60123 27.60119 27.60116 27.60114 27.60113
#> 16 26.87157 26.98208 27.01719 27.01706 27.01698 27.01693 27.01690 27.01688
#> 17 27.38151 28.45982 28.80373 28.80377 28.80380 28.80381 28.80382 28.80383
#> 18 14.89865 16.17763 16.58585 16.58623 16.58646 16.58661 16.58670 16.58676
#> 19 20.79158 21.30882 21.47385 21.47394 21.47400 21.47403 21.47405 21.47407
#> 20 14.57333 15.23489 15.44612 15.44638 15.44655 15.44665 15.44672 15.44676
#> 21 26.35106 25.47370 25.19361 25.19331 25.19312 25.19300 25.19293 25.19288
#>      0.5902   0.5505   0.5134   0.4788   0.4465   0.4164   0.3883   0.3622
#> 1  21.90158 21.90159 21.90159 21.90160 21.89776 21.84145 21.78887 21.74000
#> 2  21.17482 21.17483 21.17483 21.17483 21.17076 21.10994 21.05313 21.00035
#> 3  26.50386 26.50385 26.50385 26.50384 26.50717 26.55538 26.60036 26.64220
#> 4  19.50756 19.50755 19.50755 19.50755 19.51645 19.64968 19.77414 19.88975
#> 5  15.44678 15.44680 15.44681 15.44681 15.44873 15.48080 15.51084 15.53866
#> 6  24.13834 24.13832 24.13831 24.13830 24.13632 24.10173 24.06930 24.03931
#> 7  19.56456 19.56456 19.56455 19.56455 19.55993 19.48913 19.42295 19.36152
#> 8  14.99078 14.99079 14.99080 14.99080 14.99708 15.09445 15.18550 15.26998
#> 9  10.16284 10.16282 10.16281 10.16281 10.16295 10.16227 10.16156 10.16098
#> 10 10.38799 10.38798 10.38797 10.38796 10.38818 10.38889 10.38950 10.39013
#> 11 26.84587 26.84586 26.84585 26.84585 26.84930 26.89962 26.94660 26.99027
#> 12 27.88613 27.88612 27.88612 27.88612 27.88992 27.94669 27.99972 28.04898
#> 13 15.58928 15.58930 15.58931 15.58932 15.60030 15.76952 15.92770 16.07451
#> 14 14.67728 14.67729 14.67729 14.67730 14.67896 14.70626 14.73182 14.75550
#> 15 27.60112 27.60112 27.60112 27.60111 27.60481 27.65982 27.71119 27.75892
#> 16 27.01687 27.01686 27.01686 27.01685 27.01585 26.99910 26.98342 26.96889
#> 17 28.80383 28.80384 28.80384 28.80384 28.80343 28.79776 28.79248 28.78755
#> 18 16.58679 16.58681 16.58683 16.58683 16.58913 16.62827 16.66495 16.69889
#> 19 21.47408 21.47408 21.47409 21.47409 21.46109 21.26586 21.08348 20.91408
#> 20 15.44678 15.44680 15.44681 15.44681 15.43069 15.19024 14.96565 14.75700
#> 21 25.19285 25.19284 25.19282 25.19282 25.19120 25.16314 25.13685 25.11252
#>      0.3378   0.3150   0.2938   0.2740   0.2555   0.2383   0.2222   0.2072
#> 1  21.69426 21.65160 21.61181 21.57471 21.54393 22.09952 22.09954 22.09956
#> 2  20.95093 20.90483 20.86185 20.82176 20.78853 21.82932 21.82945 21.82954
#> 3  26.68134 26.71783 26.75187 26.78362 26.80998 26.77528 26.77556 26.77575
#> 4  19.99802 20.09901 20.19319 20.28102 20.35379 20.84410 20.84406 20.84403
#> 5  15.56480 15.58918 15.61192 15.63313 15.65057 14.47998 14.47958 14.47930
#> 6  24.01108 23.98475 23.96020 23.93730 23.91851 24.19940 24.19916 24.19899
#> 7  19.30395 19.25024 19.20016 19.15345 19.11482 18.10521 18.10481 18.10452
#> 8  15.34919 15.42307 15.49197 15.55623 15.60934 16.25547 16.25527 16.25513
#> 9  10.16036 10.15978 10.15924 10.15874 10.15842 10.33516 10.33549 10.33573
#> 10 10.39065 10.39113 10.39158 10.39200 10.39245 10.18275 10.18304 10.18326
#> 11 27.03114 27.06925 27.10480 27.13795 27.16546 28.07365 28.07389 28.07406
#> 12 28.09512 28.13815 28.17828 28.21570 28.24671 29.74425 29.74436 29.74444
#> 13 16.21213 16.34049 16.46020 16.57185 16.66417 16.67437 16.67419 16.67406
#> 14 14.77774 14.79849 14.81784 14.83589 14.85075 13.12753 13.12723 13.12701
#> 15 27.80361 27.84530 27.88417 27.92043 27.95048 28.36691 28.36705 28.36714
#> 16 26.95525 26.94252 26.93064 26.91957 26.91046 24.91040 24.91047 24.91052
#> 17 28.78296 28.77867 28.77467 28.77094 28.76784 27.21865 27.21848 27.21837
#> 18 16.73080 16.76057 16.78833 16.81423 16.83550 18.21443 18.21464 18.21479
#> 19 20.75542 20.60744 20.46942 20.34071 20.23410 19.64321 19.64306 19.64295
#> 20 14.56163 14.37940 14.20946 14.05097 13.91962 15.07267 15.07258 15.07252
#> 21 25.08964 25.06829 25.04838 25.02981 25.01457 24.74775 24.74809 24.74833
#>      0.1933   0.1803   0.1681   0.1568   0.1462   0.1364   0.1272   0.1186
#> 1  22.09957 22.09958 22.09959 22.09960 22.09960 22.09937 22.08994 22.08114
#> 2  21.82961 21.82965 21.82969 21.82971 21.82973 21.82978 21.82092 21.81245
#> 3  26.77589 26.77599 26.77607 26.77612 26.77615 26.77504 26.74591 26.71843
#> 4  20.84400 20.84399 20.84398 20.84397 20.84396 20.83716 20.71281 20.59790
#> 5  14.47910 14.47895 14.47885 14.47878 14.47872 14.47743 14.47048 14.46486
#> 6  24.19887 24.19878 24.19872 24.19868 24.19865 24.20128 24.25892 24.31266
#> 7  18.10432 18.10417 18.10406 18.10399 18.10393 18.10815 18.20298 18.29147
#> 8  16.25502 16.25495 16.25490 16.25486 16.25483 16.25208 16.20984 16.17120
#> 9  10.33590 10.33602 10.33611 10.33617 10.33621 10.33578 10.31493 10.29498
#> 10 10.18341 10.18352 10.18359 10.18365 10.18369 10.18558 10.21202 10.23588
#> 11 28.07419 28.07428 28.07434 28.07438 28.07442 28.07498 28.07821 28.08072
#> 12 29.74449 29.74453 29.74456 29.74458 29.74460 29.74562 29.76187 29.77668
#> 13 16.67397 16.67391 16.67386 16.67383 16.67380 16.66822 16.57505 16.48928
#> 14 13.12686 13.12675 13.12667 13.12662 13.12658 13.13126 13.23784 13.33704
#> 15 28.36721 28.36726 28.36730 28.36732 28.36734 28.36761 28.37011 28.37216
#> 16 24.91056 24.91058 24.91060 24.91061 24.91062 24.91541 25.00873 25.09492
#> 17 27.21828 27.21822 27.21818 27.21815 27.21813 27.21489 27.16041 27.11036
#> 18 18.21489 18.21497 18.21502 18.21506 18.21509 18.22009 18.30928 18.39137
#> 19 19.64287 19.64281 19.64277 19.64275 19.64273 19.63963 19.57553 19.51651
#> 20 15.07248 15.07244 15.07242 15.07241 15.07240 15.06976 15.00660 14.94829
#> 21 24.74850 24.74863 24.74871 24.74878 24.74882 24.75088 24.77764 24.80170
#>      0.1106   0.1031   0.0962   0.0897   0.0837   0.0780   0.0728   0.0679
#> 1  22.35931 22.35171 22.35155 22.35145 22.35138 22.35134 22.35132 22.35132
#> 2  22.13405 22.12742 22.12716 22.12697 22.12684 22.12674 22.12668 22.12665
#> 3  26.50880 26.48886 26.48869 26.48855 26.48844 26.48835 26.48829 26.48823
#> 4  20.35656 20.26377 20.26368 20.26365 20.26365 20.26366 20.26367 20.26368
#> 5  14.46581 14.45803 14.45837 14.45866 14.45890 14.45909 14.45924 14.45934
#> 6  24.61091 24.65263 24.65277 24.65283 24.65285 24.65285 24.65286 24.65287
#> 7  17.95373 18.02442 18.02495 18.02530 18.02554 18.02569 18.02578 18.02584
#> 8  16.39737 16.36345 16.36351 16.36359 16.36368 16.36376 16.36383 16.36389
#> 9  10.29145 10.27759 10.27731 10.27707 10.27689 10.27675 10.27665 10.27658
#> 10 10.21219 10.23345 10.23331 10.23316 10.23304 10.23293 10.23286 10.23280
#> 11 28.03569 28.03945 28.03931 28.03917 28.03905 28.03895 28.03888 28.03883
#> 12 29.85989 29.87240 29.87233 29.87225 29.87218 29.87212 29.87207 29.87204
#> 13 16.62059 16.54830 16.54834 16.54844 16.54856 16.54867 16.54877 16.54885
#> 14 13.29631 13.37389 13.37442 13.37480 13.37507 13.37527 13.37540 13.37549
#> 15 28.28369 28.28634 28.28630 28.28625 28.28621 28.28617 28.28614 28.28611
#> 16 25.56066 25.62846 25.62855 25.62858 25.62860 25.62862 25.62863 25.62865
#> 17 26.74981 26.70921 26.70935 26.70949 26.70961 26.70971 26.70977 26.70981
#> 18 18.34876 18.41631 18.41637 18.41639 18.41639 18.41637 18.41635 18.41632
#> 19 19.57925 19.53137 19.53120 19.53113 19.53110 19.53109 19.53109 19.53110
#> 20 14.73427 14.68916 14.68893 14.68880 14.68873 14.68867 14.68863 14.68859
#> 21 24.54091 24.56375 24.56361 24.56346 24.56332 24.56319 24.56309 24.56301
#>      0.0633   0.0590   0.0550   0.0513   0.0479   0.0446   0.0416   0.0388
#> 1  22.35131 22.35132 22.35132 22.35133 22.35134 22.35137 22.31538 22.31532
#> 2  22.12662 22.12662 22.12661 22.12661 22.12662 22.12661 22.06225 22.06213
#> 3  26.48819 26.48816 26.48814 26.48812 26.48811 26.48822 26.56979 26.56977
#> 4  20.26369 20.26370 20.26371 20.26371 20.26371 20.26387 20.22710 20.22702
#> 5  14.45942 14.45947 14.45950 14.45953 14.45954 14.45963 14.48352 14.48365
#> 6  24.65288 24.65290 24.65292 24.65294 24.65295 24.65267 24.65994 24.66009
#> 7  18.02587 18.02589 18.02590 18.02590 18.02590 18.02560 18.01458 18.01479
#> 8  16.36394 16.36398 16.36401 16.36403 16.36405 16.36404 16.19316 16.19302
#> 9  10.27653 10.27649 10.27647 10.27645 10.27644 10.27655 10.39696 10.39696
#> 10 10.23275 10.23272 10.23269 10.23268 10.23267 10.23269 10.34894 10.34897
#> 11 28.03879 28.03876 28.03874 28.03873 28.03872 28.03861 27.94966 27.94952
#> 12 29.87202 29.87201 29.87200 29.87200 29.87200 29.87176 29.64216 29.64193
#> 13 16.54891 16.54896 16.54899 16.54902 16.54903 16.54924 16.46788 16.46778
#> 14 13.37555 13.37559 13.37561 13.37562 13.37563 13.37551 13.40473 13.40491
#> 15 28.28609 28.28608 28.28607 28.28606 28.28605 28.28599 28.22501 28.22492
#> 16 25.62866 25.62868 25.62869 25.62870 25.62871 25.62884 25.95952 25.95985
#> 17 26.70984 26.70985 26.70986 26.70986 26.70986 26.71013 26.91209 26.91229
#> 18 18.41630 18.41628 18.41626 18.41624 18.41623 18.41589 18.04081 18.04043
#> 19 19.53110 19.53111 19.53112 19.53113 19.53113 19.53142 19.74065 19.74084
#> 20 14.68856 14.68854 14.68853 14.68852 14.68852 14.68855 14.58658 14.58646
#> 21 24.56295 24.56290 24.56286 24.56283 24.56281 24.56282 24.69929 24.69934
#>      0.0362   0.0338   0.0315   0.0294   0.0274   0.0256   0.0238   0.0222
#> 1  22.36238 22.36244 22.36250 22.36255 22.36261 22.36266 22.57207 22.57216
#> 2  21.97061 21.97058 21.97056 21.97054 21.97051 21.97049 22.09876 22.09877
#> 3  26.55654 26.55654 26.55653 26.55653 26.55653 26.55652 26.43034 26.43029
#> 4  20.12010 20.12003 20.11995 20.11988 20.11981 20.11974 20.12185 20.12181
#> 5  15.03296 15.03328 15.03360 15.03390 15.03419 15.03447 15.03927 15.03947
#> 6  24.90518 24.90534 24.90550 24.90565 24.90579 24.90593 24.83882 24.83890
#> 7  17.52734 17.52705 17.52677 17.52650 17.52624 17.52599 17.70824 17.70811
#> 8  15.83494 15.83471 15.83448 15.83427 15.83406 15.83387 15.59794 15.59773
#> 9  10.55948 10.55960 10.55971 10.55982 10.55993 10.56003 10.72259 10.72271
#> 10 10.33605 10.33607 10.33608 10.33609 10.33610 10.33611 10.32160 10.32159
#> 11 27.88766 27.88758 27.88751 27.88744 27.88737 27.88730 27.83747 27.83741
#> 12 29.87893 29.87899 29.87904 29.87910 29.87915 29.87919 29.91344 29.91351
#> 13 16.32149 16.32139 16.32129 16.32119 16.32110 16.32100 16.39543 16.39540
#> 14 13.56250 13.56258 13.56266 13.56274 13.56281 13.56288 13.45453 13.45453
#> 15 28.31148 28.31150 28.31151 28.31153 28.31154 28.31155 28.34762 28.34765
#> 16 25.81494 25.81496 25.81498 25.81500 25.81502 25.81504 25.74595 25.74590
#> 17 27.05429 27.05441 27.05452 27.05463 27.05473 27.05484 27.16141 27.16151
#> 18 17.96507 17.96483 17.96460 17.96438 17.96417 17.96397 17.99186 17.99179
#> 19 19.47361 19.47359 19.47356 19.47354 19.47351 19.47349 19.46133 19.46128
#> 20 14.76350 14.76360 14.76370 14.76379 14.76388 14.76396 14.59648 14.59651
#> 21 24.66094 24.66094 24.66094 24.66095 24.66095 24.66095 24.54301 24.54295
#>      0.0207   0.0193   0.0180   0.0168   0.0157   0.0146   0.0136   0.0127
#> 1  22.57226 22.57235 22.57244 22.57253 22.57261 22.57270 22.57278 22.57286
#> 2  22.09879 22.09881 22.09882 22.09884 22.09885 22.09887 22.09888 22.09890
#> 3  26.43025 26.43020 26.43015 26.43011 26.43006 26.43002 26.42998 26.42993
#> 4  20.12178 20.12176 20.12173 20.12170 20.12167 20.12164 20.12162 20.12159
#> 5  15.03967 15.03986 15.04005 15.04023 15.04041 15.04059 15.04076 15.04093
#> 6  24.83897 24.83905 24.83912 24.83919 24.83926 24.83932 24.83939 24.83945
#> 7  17.70798 17.70786 17.70774 17.70762 17.70751 17.70739 17.70728 17.70717
#> 8  15.59752 15.59732 15.59712 15.59693 15.59674 15.59656 15.59638 15.59620
#> 9  10.72283 10.72295 10.72306 10.72318 10.72329 10.72339 10.72350 10.72360
#> 10 10.32158 10.32157 10.32156 10.32155 10.32154 10.32153 10.32152 10.32151
#> 11 27.83736 27.83730 27.83725 27.83720 27.83715 27.83710 27.83705 27.83700
#> 12 29.91358 29.91364 29.91370 29.91377 29.91383 29.91389 29.91395 29.91400
#> 13 16.39538 16.39536 16.39534 16.39532 16.39530 16.39529 16.39527 16.39525
#> 14 13.45452 13.45452 13.45452 13.45452 13.45451 13.45451 13.45451 13.45451
#> 15 28.34767 28.34770 28.34772 28.34775 28.34777 28.34780 28.34782 28.34784
#> 16 25.74585 25.74581 25.74576 25.74572 25.74568 25.74564 25.74559 25.74555
#> 17 27.16162 27.16172 27.16183 27.16192 27.16202 27.16212 27.16221 27.16230
#> 18 17.99172 17.99165 17.99158 17.99151 17.99145 17.99138 17.99132 17.99126
#> 19 19.46124 19.46119 19.46115 19.46111 19.46106 19.46102 19.46098 19.46094
#> 20 14.59653 14.59655 14.59658 14.59660 14.59662 14.59664 14.59666 14.59668
#> 21 24.54289 24.54283 24.54278 24.54272 24.54267 24.54262 24.54256 24.54251
#>      0.0119   0.0111   0.0103   0.0096   0.0090   0.0084   0.0078   0.0073
#> 1  22.57294 22.57301 22.57309 22.57316 22.57323 22.57330 22.57337 22.57344
#> 2  22.09891 22.09893 22.09894 22.09895 22.09897 22.09898 22.09899 22.09901
#> 3  26.42989 26.42985 26.42981 26.42978 26.42974 26.42970 26.42967 26.42963
#> 4  20.12157 20.12154 20.12152 20.12149 20.12147 20.12145 20.12143 20.12140
#> 5  15.04109 15.04125 15.04141 15.04157 15.04172 15.04186 15.04201 15.04215
#> 6  24.83951 24.83957 24.83963 24.83969 24.83975 24.83981 24.83986 24.83991
#> 7  17.70707 17.70696 17.70686 17.70676 17.70667 17.70657 17.70648 17.70639
#> 8  15.59603 15.59586 15.59569 15.59553 15.59537 15.59522 15.59507 15.59492
#> 9  10.72370 10.72380 10.72390 10.72399 10.72408 10.72417 10.72426 10.72435
#> 10 10.32150 10.32150 10.32149 10.32148 10.32147 10.32146 10.32146 10.32145
#> 11 27.83696 27.83691 27.83687 27.83682 27.83678 27.83674 27.83670 27.83666
#> 12 29.91406 29.91411 29.91417 29.91422 29.91427 29.91432 29.91437 29.91442
#> 13 16.39523 16.39521 16.39520 16.39518 16.39516 16.39515 16.39513 16.39512
#> 14 13.45450 13.45450 13.45450 13.45450 13.45450 13.45449 13.45449 13.45449
#> 15 28.34786 28.34788 28.34790 28.34792 28.34794 28.34796 28.34798 28.34800
#> 16 25.74552 25.74548 25.74544 25.74540 25.74537 25.74533 25.74530 25.74527
#> 17 27.16239 27.16248 27.16256 27.16264 27.16273 27.16280 27.16288 27.16296
#> 18 17.99120 17.99114 17.99108 17.99102 17.99097 17.99091 17.99086 17.99081
#> 19 19.46090 19.46086 19.46083 19.46079 19.46076 19.46072 19.46069 19.46065
#> 20 14.59671 14.59672 14.59674 14.59676 14.59678 14.59680 14.59682 14.59684
#> 21 24.54246 24.54242 24.54237 24.54232 24.54228 24.54223 24.54219 24.54215
#>      0.0068   0.0063   0.0059   0.0055
#> 1  22.57350 22.57357 22.57363 22.57369
#> 2  22.09902 22.09903 22.09904 22.09905
#> 3  26.42960 26.42956 26.42953 26.42950
#> 4  20.12138 20.12136 20.12134 20.12132
#> 5  15.04229 15.04242 15.04256 15.04269
#> 6  24.83997 24.84002 24.84007 24.84012
#> 7  17.70630 17.70621 17.70612 17.70604
#> 8  15.59477 15.59463 15.59449 15.59436
#> 9  10.72443 10.72452 10.72460 10.72468
#> 10 10.32144 10.32144 10.32143 10.32142
#> 11 27.83662 27.83658 27.83654 27.83650
#> 12 29.91447 29.91451 29.91456 29.91460
#> 13 16.39510 16.39509 16.39507 16.39506
#> 14 13.45449 13.45449 13.45449 13.45448
#> 15 28.34802 28.34804 28.34805 28.34807
#> 16 25.74523 25.74520 25.74517 25.74514
#> 17 27.16303 27.16311 27.16318 27.16325
#> 18 17.99076 17.99071 17.99066 17.99061
#> 19 19.46062 19.46059 19.46056 19.46053
#> 20 14.59685 14.59687 14.59689 14.59690
#> 21 24.54211 24.54207 24.54203 24.54199
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 22.8 18.1 14.3 22.8 17.8 17.3 10.4 14.7 32.4 33.9 15.5 13.3 27.3
#> [16] 26.0 30.4 15.8 19.7 15.0 21.4
#> 
#> $X
#>               am        carb       cyl        disp        drat       gear
#>  [1,]  0.8660254  0.53338075  0.000000 -0.45731690  0.45261337  0.1270001
#>  [2,]  0.8660254  0.53338075  0.000000 -0.45731690  0.45261337  0.1270001
#>  [3,]  0.8660254 -1.14676861 -1.145644 -0.88509663  0.34884829  0.1270001
#>  [4,] -1.1547005 -1.14676861  0.000000  0.07740776 -1.91323030 -1.2065012
#>  [5,] -1.1547005  0.53338075  1.145644  1.18798975 -0.97934464 -1.2065012
#>  [6,] -1.1547005 -0.58671882 -1.145644 -0.61526634  0.49411940  0.1270001
#>  [7,] -1.1547005  0.53338075  0.000000 -0.39479525  0.49411940  0.1270001
#>  [8,] -1.1547005 -0.02666904  1.145644  0.49531565 -1.26988685 -1.2065012
#>  [9,] -1.1547005  0.53338075  1.145644  2.01064307 -1.41515795 -1.2065012
#> [10,] -1.1547005  0.53338075  1.145644  1.84611241 -0.93783861 -1.2065012
#> [11,]  0.8660254 -1.14676861 -1.145644 -1.12613405  0.82616763  0.1270001
#> [12,]  0.8660254 -1.14676861 -1.145644 -1.18865571  1.11670983  0.1270001
#> [13,] -1.1547005 -0.58671882  1.145644  0.84247535 -1.91323030 -1.2065012
#> [14,] -1.1547005  0.53338075  1.145644  1.10572442  0.09981212 -1.2065012
#> [15,]  0.8660254 -1.14676861 -1.145644 -1.12366609  0.82616763  0.1270001
#> [16,]  0.8660254 -0.58671882 -1.145644 -0.78391027  1.55252314  1.4605015
#> [17,]  0.8660254 -0.58671882 -1.145644 -0.99121891  0.18282418  1.4605015
#> [18,]  0.8660254  0.53338075  1.145644  1.11395095  1.11670983  1.4605015
#> [19,]  0.8660254  1.65348032  0.000000 -0.58071490 -0.12847104  1.4605015
#> [20,]  0.8660254  2.77357988  1.145644  0.70262429 -0.29449516  1.4605015
#> [21,]  0.8660254 -0.58671882 -1.145644 -0.77815170  0.88842667  0.1270001
#>                hp        qsec         vs          wt
#>  [1,] -0.55712914 -0.58428162 -0.8660254 -0.49020959
#>  [2,] -0.55712914 -0.30359731 -0.8660254 -0.23252113
#>  [3,] -0.78443783  0.49334564  1.1547005 -0.79337249
#>  [4,] -0.62398463  1.30031303  1.1547005  0.35864652
#>  [5,]  1.24796927 -0.89503925 -0.8660254  0.46980625
#>  [6,] -0.75769563  2.64358794  1.1547005  0.04537819
#>  [7,] -0.38330485  0.63870001  1.1547005  0.33843566
#>  [8,]  0.37884781 -0.01288857 -0.8660254  0.63149313
#>  [9,]  0.84683629  0.09738027 -0.8660254  2.34335296
#> [10,]  1.04740278 -0.10310852 -0.8660254  2.26352007
#> [11,] -1.14545751  0.92439654  1.1547005 -0.91463765
#> [12,] -1.15882861  1.13992199  1.1547005 -1.28348585
#> [13,] -0.02228517 -0.37878061 -0.8660254  0.41927910
#> [14,]  1.24796927 -1.11056471 -0.8660254  0.74265286
#> [15,] -1.14545751  0.63870001  1.1547005 -1.18243155
#> [16,] -0.81118002 -0.46398835 -0.8660254 -0.97527023
#> [17,] -0.51701584 -0.36374395  1.1547005 -1.60888069
#> [18,]  1.50202015 -1.56667671 -0.8660254  0.06558905
#> [19,]  0.31199232 -1.06545473 -0.8660254 -0.33862815
#> [20,]  2.45136820 -1.51655451 -0.8660254  0.46980625
#> [21,] -0.57050024  0.48833342  1.1547005 -0.32852272
#> attr(,"center")
#>  [1]   0.5714286   3.0476190   6.0000000 215.5904762   3.6819048   3.9047619
#>  [7] 151.6666667  17.6257143   0.4285714   3.1050952
#> attr(,"scale")
#>  [1]   0.4948717   1.7855555   1.7457431 121.5578871   0.4818577   0.7499055
#>  [7]  74.7881665   1.9951240   0.4948717   0.9895670
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 38
#> 
#> $lambda.min
#> [1] 0.4164054
#> 
#> $null.dev
#> [1] 39.75011
#> 
#> $Bias
#> [1] 6.883982
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 5.344532