Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 47.393968 45.949010 40.212220 33.946515 28.728484 24.406061 20.848036
#>   [8] 17.941274 15.588287 13.705140 12.219626 11.069684 10.202031  9.570967
#>  [15]  9.137348  8.863460  8.678938  8.643712  8.668871  8.668871  8.716830
#>  [22]  8.799845  8.882786  8.970172  8.972528  8.942670  8.917498  8.874800
#>  [29]  8.838212  8.781196  8.748248  8.792654  9.436581  9.802366 10.135708
#>  [36] 10.345737 10.357393 10.305643 10.349801 10.300022 10.127141 20.014074
#>  [43] 19.494608 20.016749 20.019972 22.525383 22.512028 23.961253 24.319843
#>  [50] 25.288050 33.024521 39.554331 39.612299 40.912369 41.119726 41.115933
#>  [57] 41.348867 41.352032 41.312317 42.662385 42.876483 42.879577 42.621486
#>  [64] 42.391013 42.393544 42.175677 42.148359 42.152033 41.940283 42.228754
#>  [71] 42.232499 42.128698 43.395464 43.334827 43.335589 43.070066 43.070872
#>  [78] 43.071569 43.072111 42.762412 42.761720 42.761068 42.760450 42.759860
#>  [85] 42.759292 42.758742 42.758206 42.757680 42.757162 42.601734 42.600933
#>  [92] 42.600147 42.599373 42.598609 42.597855 42.597109 42.596370 42.595638
#>  [99] 42.594913 42.594194
#> 
#> $cvse
#>   [1] 12.166344 12.078167 10.543282  8.893842  7.602216  6.603039  5.835107
#>   [8]  5.242602  4.777174  4.399694  4.080814  3.800274  3.545420  3.309506
#>  [15]  3.090152  2.887950  2.708352  2.613390  2.611403  2.611403  2.612032
#>  [22]  2.613949  2.616930  2.620159  2.632942  2.628215  2.611811  2.570883
#>  [29]  2.531343  2.492878  2.457952  2.423715  2.423326  2.403939  2.396135
#>  [36]  2.404758  2.415325  2.400518  2.361022  2.342141  2.329972  7.712585
#>  [43]  7.669357  7.613724  7.607167  7.755490  7.757149  7.827530  7.784883
#>  [50]  7.770024 13.370731 16.331203 16.316540 16.245136 16.251848 16.255245
#>  [57] 16.253606 16.255973 16.269514 16.392898 16.417827 16.419775 16.448313
#>  [64] 16.466503 16.468223 16.617061 16.622093 16.624492 16.648764 16.622760
#>  [71] 16.624855 16.637015 17.332208 17.331688 17.331772 17.309628 17.309610
#>  [78] 17.309538 17.309400 17.225366 17.224893 17.224408 17.223910 17.223401
#>  [85] 17.222882 17.222353 17.221815 17.221269 17.220715 17.222551 17.221990
#>  [92] 17.221423 17.220851 17.220275 17.219696 17.219115 17.218532 17.217947
#>  [99] 17.217363 17.216779
#> 
#> $fold
#>  [1]  9  2  7  1 10  9  3  2  6  8  1  5  8  7 10  1  4  6  3  5  4
#> 
#> $lambda
#>   [1] 5.849379456 5.455150858 5.087491948 4.744612018 4.424841048 4.126621570
#>   [7] 3.848501088 3.589124995 3.347229983 3.121637885 2.911249939 2.715041437
#>  [13] 2.532056731 2.361404582 2.202253816 2.053829280 1.915408060 1.786315968
#>  [19] 1.665924250 1.553646531 1.448935955 1.351282521 1.260210602 1.175276625
#>  [25] 1.096066914 1.022195672 0.953303105 0.889053666 0.829134425 0.773253540
#>  [31] 0.721138839 0.672536495 0.627209786 0.584937946 0.545515086 0.508749195
#>  [37] 0.474461202 0.442484105 0.412662157 0.384850109 0.358912500 0.334722998
#>  [43] 0.312163788 0.291124994 0.271504144 0.253205673 0.236140458 0.220225382
#>  [49] 0.205382928 0.191540806 0.178631597 0.166592425 0.155364653 0.144893595
#>  [55] 0.135128252 0.126021060 0.117527662 0.109606691 0.102219567 0.095330311
#>  [61] 0.088905367 0.082913442 0.077325354 0.072113884 0.067253649 0.062720979
#>  [67] 0.058493795 0.054551509 0.050874919 0.047446120 0.044248410 0.041266216
#>  [73] 0.038485011 0.035891250 0.033472300 0.031216379 0.029112499 0.027150414
#>  [79] 0.025320567 0.023614046 0.022022538 0.020538293 0.019154081 0.017863160
#>  [85] 0.016659242 0.015536465 0.014489360 0.013512825 0.012602106 0.011752766
#>  [91] 0.010960669 0.010221957 0.009533031 0.008890537 0.008291344 0.007732535
#>  [97] 0.007211388 0.006725365 0.006272098 0.005849379
#> 
#> $fit
#> $beta
#>             5.8494    5.4552   5.0875   4.7446   4.4248    4.1266    3.8485
#> (Intercept)   21.8 23.518400 25.12099 26.61556 28.00941 29.309317 30.521615
#> am             0.0  0.000000  0.00000  0.00000  0.00000  0.000000  0.000000
#> carb           0.0  0.000000  0.00000  0.00000  0.00000  0.000000  0.000000
#> cyl            0.0  0.000000  0.00000  0.00000  0.00000  0.000000  0.000000
#> disp           0.0  0.000000  0.00000  0.00000  0.00000  0.000000  0.000000
#> drat           0.0  0.000000  0.00000  0.00000  0.00000  0.000000  0.000000
#> gear           0.0  0.000000  0.00000  0.00000  0.00000  0.000000  0.000000
#> hp             0.0  0.000000  0.00000  0.00000  0.00000  0.000000  0.000000
#> qsec           0.0  0.000000  0.00000  0.00000  0.00000  0.000000  0.000000
#> vs             0.0  0.000000  0.00000  0.00000  0.00000  0.000000  0.000000
#> wt             0.0 -0.575752 -1.11270 -1.61346 -2.08047 -2.516005 -2.922187
#>                3.5891    3.3472    3.1216    2.9112   2.7150    2.5321   2.3614
#> (Intercept) 31.652208 32.706602 33.689934 34.606992 35.46224 36.259855 37.00371
#> am           0.000000  0.000000  0.000000  0.000000  0.00000  0.000000  0.00000
#> carb         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000  0.00000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.00000  0.000000  0.00000
#> disp         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000  0.00000
#> drat         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000  0.00000
#> gear         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000  0.00000
#> hp           0.000000  0.000000  0.000000  0.000000  0.00000  0.000000  0.00000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000  0.00000
#> vs           0.000000  0.000000  0.000000  0.000000  0.00000  0.000000  0.00000
#> wt          -3.300993 -3.654269 -3.983736 -4.290997 -4.57755 -4.844791 -5.09402
#>                2.2023    2.0538    1.9154    1.7863    1.6659    1.5536
#> (Intercept) 37.697431 38.344397 38.797880 38.797880 38.797880 38.797880
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -5.326452 -5.543219 -5.695159 -5.695159 -5.695159 -5.695159
#>                1.4489    1.3513    1.2602    1.1753    1.0961      1.0222
#> (Intercept) 38.797880 38.797880 38.797880 38.797880 38.399128 37.03426191
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.021147  0.09352987
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> wt          -5.695159 -5.695159 -5.695159 -5.695159 -5.692298 -5.68250436
#>                 0.9533    0.8891     0.8291     0.7733    0.7211     0.6725
#> (Intercept) 35.7614352 34.574393 33.4673530 32.4349240 31.472077 30.5741232
#> am           0.0000000  0.000000  0.0000000  0.0000000  0.000000  0.0000000
#> carb         0.0000000  0.000000  0.0000000  0.0000000  0.000000  0.0000000
#> cyl          0.0000000  0.000000  0.0000000  0.0000000  0.000000  0.0000000
#> disp         0.0000000  0.000000  0.0000000  0.0000000  0.000000  0.0000000
#> drat         0.0000000  0.000000  0.0000000  0.0000000  0.000000  0.0000000
#> gear         0.0000000  0.000000  0.0000000  0.0000000  0.000000  0.0000000
#> hp           0.0000000  0.000000  0.0000000  0.0000000  0.000000  0.0000000
#> qsec         0.1610316  0.223984  0.2826936  0.3374464  0.388509  0.4361301
#> vs           0.0000000  0.000000  0.0000000  0.0000000  0.000000  0.0000000
#> wt          -5.6733714 -5.664854 -5.6569106 -5.6495026 -5.642594 -5.6361507
#>                 0.6272     0.5849     0.5455     0.5087     0.4745     0.4425
#> (Intercept) 29.7366883 28.9556936 25.7755322 22.7624815 19.9516486 17.3348114
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.4157934  0.8114663  1.1806110  1.5241346
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.4805418  0.5219603  0.5804777  0.6354623  0.6867496  0.7345352
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.6301419 -5.6245380 -5.4382496 -5.2610707 -5.0957724 -4.9419391
#>                 0.4127     0.3849    0.3589    0.3347    0.3122     0.2911
#> (Intercept) 14.8898186 14.2822401 14.515317 14.513358 14.512099 14.1564762
#> am           0.0000000  0.0000000  3.240438  3.240998  3.241358  3.1778549
#> carb         0.0000000  0.0000000  0.000000  0.000000  0.000000  0.0000000
#> cyl          0.0000000  0.0000000  0.000000  0.000000  0.000000  0.0000000
#> disp         0.0000000  0.0000000  0.000000  0.000000  0.000000  0.0000000
#> drat         0.0000000  0.0000000  0.000000  0.000000  0.000000  0.0000000
#> gear         1.8452423  1.9314419  0.000000  0.000000  0.000000  0.1005818
#> hp           0.0000000  0.0000000  0.000000  0.000000  0.000000  0.0000000
#> qsec         0.7791449  0.7885339  1.020412  1.020469  1.020507  1.0179635
#> vs           0.0000000  0.0000000  0.000000  0.000000  0.000000  0.0000000
#> wt          -4.7981516 -4.7599025 -4.384951 -4.384741 -4.384606 -4.3647716
#>                 0.2715     0.2532     0.2361     0.2202     0.2054     0.1915
#> (Intercept) 12.5483386 11.0475040 10.4583458 10.3733902 10.3359528 10.3351872
#> am           2.8688375  2.5800117  2.4655578  2.5384439  2.5629921  2.5625592
#> carb         0.0000000  0.0000000  0.0000000 -0.4477056 -0.5900239 -0.5902789
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.5682406  1.0049446  1.1769042  1.5895455  1.7216583  1.7222772
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         1.0048703  0.9926199  0.9877593  0.8848254  0.8522946  0.8522114
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -4.2777017 -4.1964932 -4.1647828 -3.6893573 -3.5372992 -3.5370387
#>                 0.1786     0.1666     0.1554     0.1449     0.1351     0.1260
#> (Intercept) 10.3346692 10.3343260 10.3341049 10.3339676 10.3338871 -15.142123
#> am           2.5622133  2.5619396  2.5617249  2.5615578  2.5614287   2.697063
#> carb        -0.5904662 -0.5906028 -0.5907016 -0.5907723 -0.5908225  -1.283409
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000   1.272146
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000   0.000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000   2.069421
#> gear         1.7227402  1.7230841  1.7233378  1.7235235  1.7236583   2.941000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000   0.000000
#> qsec         0.8521481  0.8521002  0.8520642  0.8520373  0.8520174   1.236708
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000   0.000000
#> wt          -3.5368540 -3.5367248 -3.5366354 -3.5365746 -3.5365341  -3.303677
#>                 0.1175     0.1096     0.1022     0.0953     0.0889     0.0829
#> (Intercept) -15.147291 -15.151711 -15.155483 -15.158696 -15.161426 -15.163742
#> am            2.697737   2.698326   2.698842   2.699291   2.699683   2.700023
#> carb         -1.283420  -1.283428  -1.283432  -1.283434  -1.283433  -1.283432
#> cyl           1.272441   1.272694   1.272910   1.273095   1.273252   1.273385
#> disp          0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> drat          2.069108   2.068831   2.068585   2.068369   2.068177   2.068009
#> gear          2.941232   2.941428   2.941592   2.941729   2.941843   2.941938
#> hp            0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> qsec          1.236923   1.237110   1.237271   1.237411   1.237531   1.237634
#> vs            0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> wt           -3.303828  -3.303960  -3.304074  -3.304173  -3.304260  -3.304334
#>                 0.0773       0.0721       0.0673       0.0627       0.0585
#> (Intercept) -15.165701 -12.49197229 -12.48799836 -12.48415386 -12.48043851
#> am            2.700319   2.82786892   2.82787569   2.82787750   2.82787468
#> carb         -1.283429  -0.97523110  -0.97501858  -0.97481670  -0.97462518
#> cyl           1.273498   1.27867021   1.27832141   1.27798458   1.27765965
#> disp          0.000000   0.01581787   0.01581923   0.01582043   0.01582146
#> drat          2.067861   1.84258696   1.84224757   1.84192532   1.84161976
#> gear          2.942016   3.07007150   3.06966411   3.06927024   3.06888985
#> hp            0.000000  -0.02712478  -0.02712010  -0.02711546  -0.02711089
#> qsec          1.237722   1.13815126   1.13819430   1.13823410   1.13827080
#> vs            0.000000   1.05939314   1.05905935   1.05873233   1.05841233
#> wt           -3.304399  -3.90180693  -3.90220028  -3.90256965  -3.90291583
#>                   0.0546       0.0509       0.0474       0.0442       0.0413
#> (Intercept) -12.47685176 -12.47339284 -12.47006072 -12.46685416 -12.46377175
#> am            2.82786755   2.82785642   2.82784160   2.82782336   2.82780197
#> carb         -0.97444374  -0.97427210  -0.97410998  -0.97395708  -0.97381312
#> cyl           1.27734654   1.27704514   1.27675532   1.27647693   1.27620982
#> disp          0.01582235   0.01582308   0.01582368   0.01582415   0.01582450
#> drat          1.84133044   1.84105689   1.84079865   1.84055524   1.84032620
#> gear          3.06852287   3.06816920   3.06782871   3.06750127   3.06718671
#> hp           -0.02710639  -0.02710197  -0.02709763  -0.02709336  -0.02708919
#> qsec          1.13830453   1.13833541   1.13836356   1.13838913   1.13841221
#> vs            1.05809959   1.05779430   1.05749666   1.05720678   1.05692479
#> wt           -3.90323964  -3.90354189  -3.90382337  -3.90408488  -3.90432718
#>                   0.0385       0.0359       0.0335       0.0312       0.0291
#> (Intercept) -12.46081187 -12.45797276 -12.45525249 -12.45264902 -12.45016018
#> am            2.82777771   2.82775082   2.82772154   2.82769010   2.82765671
#> carb         -0.97367781  -0.97355086  -0.97343197  -0.97332087  -0.97321728
#> cyl           1.27595381   1.27570871   1.27547433   1.27525045   1.27503686
#> disp          0.01582474   0.01582487   0.01582490   0.01582484   0.01582470
#> drat          1.84011105   1.83990933   1.83972057   1.83954430   1.83938008
#> gear          3.06688486   3.06659552   3.06631849   3.06605353   3.06580042
#> hp           -0.02708511  -0.02708112  -0.02707723  -0.02707343  -0.02706974
#> qsec          1.13843294   1.13845142   1.13846777   1.13848209   1.13849450
#> vs            1.05665078   1.05638482   1.05612694   1.05587717   1.05563552
#> wt           -3.90455106  -3.90475727  -3.90494656  -3.90511966  -3.90527729
#>                   0.0272       0.0253       0.0236       0.0220       0.0205
#> (Intercept) -12.44778368 -12.44551716 -12.44335816 -12.44130414 -12.43935253
#> am            2.82762159   2.82758492   2.82754690   2.82750771   2.82746750
#> carb         -0.97312090  -0.97303147  -0.97294870  -0.97287234  -0.97280210
#> cyl           1.27483333   1.27463963   1.27445552   1.27428076   1.27411508
#> disp          0.01582447   0.01582417   0.01582380   0.01582337   0.01582288
#> drat          1.83922743   1.83908591   1.83895508   1.83883450   1.83872373
#> gear          3.06555891   3.06532875   3.06510968   3.06490142   3.06470371
#> hp           -0.02706615  -0.02706266  -0.02705927  -0.02705599  -0.02705281
#> qsec          1.13850509   1.13851397   1.13852123   1.13852696   1.13853127
#> vs            1.05540196   1.05517647   1.05495900   1.05474948   1.05454784
#> wt           -3.90542016  -3.90554895  -3.90566433  -3.90576697  -3.90585750
#>                   0.0192       0.0179       0.0167       0.0155       0.0145
#> (Intercept) -12.43750068 -12.43574591 -12.43408550 -12.43251671 -12.43103679
#> am            2.82742644   2.82738469   2.82734237   2.82729962   2.82725657
#> carb         -0.97273773  -0.97267898  -0.97262558  -0.97257730  -0.97253389
#> cyl           1.27395826   1.27381002   1.27367012   1.27353830   1.27341429
#> disp          0.01582234   0.01582175   0.01582112   0.01582045   0.01581974
#> drat          1.83862236   1.83852997   1.83844616   1.83837051   1.83830266
#> gear          3.06451626   3.06433879   3.06417102   3.06401266   3.06386341
#> hp           -0.02704973  -0.02704676  -0.02704390  -0.02704113  -0.02703847
#> qsec          1.13853423   1.13853593   1.13853646   1.13853589   1.13853430
#> vs            1.05435399   1.05416782   1.05398924   1.05381812   1.05365433
#> wt           -3.90593655  -3.90600473  -3.90606263  -3.90611081  -3.90614984
#>                   0.0135       0.0126       0.0118       0.0110       0.0102
#> (Intercept) -12.42964297 -12.42833247 -12.42710254 -12.42595042 -12.42487335
#> am            2.82721333   2.82717001   2.82712672   2.82708355   2.82704058
#> carb         -0.97249511  -0.97246075  -0.97243056  -0.97240434  -0.97238187
#> cyl           1.27329785   1.27318872   1.27308663   1.27299135   1.27290260
#> disp          0.01581901   0.01581824   0.01581746   0.01581665   0.01581583
#> drat          1.83824221   1.83818879   1.83814205   1.83810163   1.83806719
#> gear          3.06372299   3.06359112   3.06346749   3.06335182   3.06324384
#> hp           -0.02703591  -0.02703344  -0.02703108  -0.02702881  -0.02702664
#> qsec          1.13853176   1.13852834   1.13852411   1.13851914   1.13851348
#> vs            1.05349774   1.05334821   1.05320558   1.05306970   1.05294042
#> wt           -3.90618025  -3.90620256  -3.90621728  -3.90622489  -3.90622585
#>                   0.0095       0.0089       0.0083       0.0077       0.0072
#> (Intercept) -12.42386863 -12.42293355 -12.42206544 -12.42126167 -12.42051964
#> am            2.82699790   2.82695558   2.82691369   2.82687231   2.82683147
#> carb         -0.97236295  -0.97234737  -0.97233495  -0.97232549  -0.97231882
#> cyl           1.27282016   1.27274376   1.27267318   1.27260816   1.27254847
#> disp          0.01581499   0.01581414   0.01581328   0.01581242   0.01581155
#> drat          1.83803840   1.83801494   1.83799650   1.83798277   1.83797347
#> gear          3.06314324   3.06304976   3.06296312   3.06288304   3.06280925
#> hp           -0.02702456  -0.02702258  -0.02702068  -0.02701887  -0.02701715
#> qsec          1.13850719   1.13850034   1.13849298   1.13848515   1.13847691
#> vs            1.05281757   1.05270099   1.05259051   1.05248596   1.05238716
#> wt           -3.90622063  -3.90620964  -3.90619332  -3.90617206  -3.90614624
#>                   0.0067       0.0063       0.0058
#> (Intercept) -12.41983678 -12.41921059 -12.41863859
#> am            2.82679125   2.82675168   2.82671281
#> carb         -0.97231476  -0.97231313  -0.97231379
#> cyl           1.27249389   1.27244418   1.27239913
#> disp          0.01581069   0.01580982   0.01580895
#> drat          1.83796831   1.83796703   1.83796935
#> gear          3.06274149   3.06267950   3.06262303
#> hp           -0.02701551  -0.02701396  -0.02701248
#> qsec          1.13846830   1.13845937   1.13845016
#> vs            1.05229395   1.05220616   1.05212360
#> wt           -3.90611624  -3.90608240  -3.90604507
#> 
#> $iter
#>   [1]   0   4   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1   1   3   3   3   3   3   3   3   3   3   3  21  18
#>  [37]  18  17  17   7  52   1   1  21  31  31  13 167  21   1   1   1   1   1
#>  [55]   1  59   1   1   1   1   1   1   1 247   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.849379456 5.455150858 5.087491948 4.744612018 4.424841048 4.126621570
#>   [7] 3.848501088 3.589124995 3.347229983 3.121637885 2.911249939 2.715041437
#>  [13] 2.532056731 2.361404582 2.202253816 2.053829280 1.915408060 1.786315968
#>  [19] 1.665924250 1.553646531 1.448935955 1.351282521 1.260210602 1.175276625
#>  [25] 1.096066914 1.022195672 0.953303105 0.889053666 0.829134425 0.773253540
#>  [31] 0.721138839 0.672536495 0.627209786 0.584937946 0.545515086 0.508749195
#>  [37] 0.474461202 0.442484105 0.412662157 0.384850109 0.358912500 0.334722998
#>  [43] 0.312163788 0.291124994 0.271504144 0.253205673 0.236140458 0.220225382
#>  [49] 0.205382928 0.191540806 0.178631597 0.166592425 0.155364653 0.144893595
#>  [55] 0.135128252 0.126021060 0.117527662 0.109606691 0.102219567 0.095330311
#>  [61] 0.088905367 0.082913442 0.077325354 0.072113884 0.067253649 0.062720979
#>  [67] 0.058493795 0.054551509 0.050874919 0.047446120 0.044248410 0.041266216
#>  [73] 0.038485011 0.035891250 0.033472300 0.031216379 0.029112499 0.027150414
#>  [79] 0.025320567 0.023614046 0.022022538 0.020538293 0.019154081 0.017863160
#>  [85] 0.016659242 0.015536465 0.014489360 0.013512825 0.012602106 0.011752766
#>  [91] 0.010960669 0.010221957 0.009533031 0.008890537 0.008291344 0.007732535
#>  [97] 0.007211388 0.006725365 0.006272098 0.005849379
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 34
#> 
#> $loss
#>   [1] 876.50000 738.56588 623.16347 527.05030 447.42694 381.87796 328.32055
#>   [8] 284.96003 250.25116 222.86461 201.65782 185.64979 173.99917 165.98524
#>  [15] 160.99150 158.49137 157.97996 157.97996 157.97996 157.97996 157.97996
#>  [22] 157.97996 157.97996 157.97996 156.46409 151.61414 147.56356 144.19677
#>  [29] 141.41418 139.12987 137.26979 135.77014 134.57601 133.64019 126.85090
#>  [36] 121.92402 118.66367 116.78792 116.04493 116.01129 111.25351 111.25351
#>  [43] 111.25351 110.50266 107.89252 106.76443 106.66720 103.70588 103.52109
#>  [50] 103.52108 103.52108 103.52107 103.52107 103.52107 103.52107  94.26033
#>  [57]  94.26032  94.26031  94.26030  94.26030  94.26029  94.26029  94.26029
#>  [64]  89.16278  89.16276  89.16275  89.16273  89.16272  89.16271  89.16270
#>  [71]  89.16269  89.16268  89.16267  89.16266  89.16266  89.16265  89.16265
#>  [78]  89.16264  89.16264  89.16263  89.16263  89.16263  89.16262  89.16262
#>  [85]  89.16262  89.16262  89.16261  89.16261  89.16261  89.16261  89.16261
#>  [92]  89.16261  89.16261  89.16261  89.16260  89.16260  89.16260  89.16260
#>  [99]  89.16260  89.16260
#> 
#> $linear.predictors
#>    5.8494   5.4552   5.0875   4.7446   4.4248   4.1266   3.8485   3.5891
#> 1    21.8 22.00993 22.20571 22.38830 22.55858 22.71738 22.86549 23.00361
#> 2    21.8 21.86311 21.92197 21.97687 22.02806 22.07580 22.12033 22.16185
#> 3    21.8 22.18266 22.53952 22.87234 23.18272 23.47219 23.74214 23.99390
#> 4    21.8 21.53781 21.29330 21.06526 20.85259 20.65426 20.46929 20.29679
#> 5    21.8 21.52630 21.27104 21.03299 20.81098 20.60394 20.41085 20.23077
#> 6    21.8 21.68175 21.57147 21.46863 21.37271 21.28326 21.19984 21.12204
#> 7    21.8 21.70478 21.61598 21.53316 21.45593 21.38390 21.31673 21.25408
#> 8    21.8 21.53781 21.29330 21.06526 20.85259 20.65426 20.46929 20.29679
#> 9    21.8 21.17509 20.59230 20.04878 19.54190 19.06918 18.62831 18.21716
#> 10   21.8 20.49570 19.27931 18.14490 17.08694 16.10029 15.18013 14.32199
#> 11   21.8 20.39552 19.08570 17.86416 16.72494 15.66250 14.67167 13.74762
#> 12   21.8 22.25175 22.67305 23.06595 23.43238 23.77411 24.09280 24.39002
#> 13   21.8 22.58856 23.32398 24.00983 24.64945 25.24597 25.80228 26.32110
#> 14   21.8 22.46190 23.07918 23.65486 24.19175 24.69245 25.15940 25.59488
#> 15   21.8 22.09917 22.37818 22.63838 22.88105 23.10736 23.31842 23.51526
#> 16   21.8 21.49175 21.20428 20.93618 20.68616 20.45298 20.23552 20.03271
#> 17   21.8 21.54069 21.29886 21.07333 20.86300 20.66684 20.48390 20.31330
#> 18   21.8 22.40432 22.96791 23.49352 23.98370 24.44085 24.86718 25.26479
#> 19   21.8 22.28629 22.73981 23.16276 23.55720 23.92507 24.26813 24.58808
#> 20   21.8 22.64729 23.43747 24.17440 24.86166 25.50260 26.10035 26.65780
#> 21   21.8 21.91781 22.02768 22.13014 22.22570 22.31482 22.39794 22.47545
#>      3.3472   3.1216   2.9112   2.7150    2.5321    2.3614    2.2023    2.0538
#> 1  23.13242 23.25255 23.36458 23.46906 23.566503 23.657377 23.742126 23.821163
#> 2  22.20058 22.23669 22.27038 22.30179 22.331081 22.358402 22.383881 22.407642
#> 3  24.22870 24.44767 24.65188 24.84233 25.019940 25.185583 25.340062 25.484129
#> 4  20.13592 19.98588 19.84596 19.71547 19.593775 19.480280 19.374435 19.275724
#> 5  20.06283 19.90621 19.76014 19.62392 19.496879 19.378400 19.267906 19.164859
#> 6  21.04948 20.98182 20.91871 20.85986 20.804972 20.753785 20.706048 20.661528
#> 7  21.19565 21.14117 21.09035 21.04296 20.998764 20.957546 20.919106 20.883257
#> 8  20.13592 19.98588 19.84596 19.71547 19.593775 19.480280 19.374435 19.275724
#> 9  17.83373 17.47613 17.14263 16.83161 16.541556 16.271048 16.018770 15.783496
#> 10 13.52169 12.77532 12.07926 11.43010 10.824703 10.260104  9.733557  9.242497
#> 11 12.88584 12.08215 11.33262 10.63361  9.981709  9.373744  8.806754  8.277977
#> 12 24.66721 24.92572 25.16680 25.39163 25.601315 25.796865 25.979236 26.149315
#> 13 26.80496 27.25620 27.67703 28.06950 28.435518 28.776867 29.095210 29.392098
#> 14 26.00102 26.37978 26.73301 27.06244 27.369664 27.656183 27.923391 28.172590
#> 15 23.69883 23.87003 24.02968 24.17858 24.317446 24.446950 24.567726 24.680362
#> 16 19.84357 19.66718 19.50268 19.34927 19.206191 19.072759 18.948319 18.832266
#> 17 20.15419 20.00580 19.86742 19.73836 19.617998 19.505750 19.401067 19.303440
#> 18 25.63559 25.98141 26.30391 26.60468 26.885185 27.146781 27.390746 27.618268
#> 19 24.88647 25.16474 25.42426 25.66629 25.892003 26.102506 26.298823 26.481908
#> 20 27.17769 27.66254 28.11471 28.53641 28.929686 29.296457 29.638509 29.957507
#> 21 22.54773 22.61515 22.67802 22.73665 22.791336 22.842334 22.889894 22.934248
#>       1.9154    1.7863    1.6659    1.5536    1.4489    1.3513    1.2602
#> 1  23.876563 23.876563 23.876563 23.876563 23.876563 23.876563 23.876563
#> 2  22.424298 22.424298 22.424298 22.424298 22.424298 22.424298 22.424298
#> 3  25.585111 25.585111 25.585111 25.585111 25.585111 25.585111 25.585111
#> 4  19.206533 19.206533 19.206533 19.206533 19.206533 19.206533 19.206533
#> 5  19.092630 19.092630 19.092630 19.092630 19.092630 19.092630 19.092630
#> 6  20.630323 20.630323 20.630323 20.630323 20.630323 20.630323 20.630323
#> 7  20.858129 20.858129 20.858129 20.858129 20.858129 20.858129 20.858129
#> 8  19.206533 19.206533 19.206533 19.206533 19.206533 19.206533 19.206533
#> 9  15.618583 15.618583 15.618583 15.618583 15.618583 15.618583 15.618583
#> 10  8.898296  8.898296  8.898296  8.898296  8.898296  8.898296  8.898296
#> 11  7.907338  7.907338  7.907338  7.907338  7.907338  7.907338  7.907338
#> 12 26.268530 26.268530 26.268530 26.268530 26.268530 26.268530 26.268530
#> 13 29.600198 29.600198 29.600198 29.600198 29.600198 29.600198 29.600198
#> 14 28.347263 28.347263 28.347263 28.347263 28.347263 28.347263 28.347263
#> 15 24.759313 24.759313 24.759313 24.759313 24.759313 24.759313 24.759313
#> 16 18.750920 18.750920 18.750920 18.750920 18.750920 18.750920 18.750920
#> 17 19.235009 19.235009 19.235009 19.235009 19.235009 19.235009 19.235009
#> 18 27.777747 27.777747 27.777747 27.777747 27.777747 27.777747 27.777747
#> 19 26.610240 26.610240 26.610240 26.610240 26.610240 26.610240 26.610240
#> 20 30.181104 30.181104 30.181104 30.181104 30.181104 30.181104 30.181104
#> 21 22.965338 22.965338 22.965338 22.965338 22.965338 22.965338 22.965338
#>       1.1753    1.0961    1.0222    0.9533    0.8891    0.8291    0.7733
#> 1  23.876563 23.833387 23.685602 23.547783 23.419252 23.299384 23.187595
#> 2  22.424298 22.393694 22.288940 22.191251 22.100145 22.015180 21.935941
#> 3  25.585111 25.586543 25.591443 25.596012 25.600274 25.604248 25.607955
#> 4  19.206533 19.177545 19.078325 18.985796 18.899503 18.819026 18.743972
#> 5  19.092630 19.131370 19.263971 19.387630 19.502955 19.610507 19.710811
#> 6  20.630323 20.663638 20.777670 20.884013 20.983189 21.075680 21.161938
#> 7  20.858129 20.952656 21.276207 21.577940 21.859337 22.121768 22.366513
#> 8  19.206533 19.217302 19.254162 19.288536 19.320593 19.350490 19.378372
#> 9  15.618583 15.599434 15.533889 15.472764 15.415759 15.362596 15.313015
#> 10  8.898296  8.894788  8.882781  8.871584  8.861142  8.851403  8.842321
#> 11  7.907338  7.900944  7.879061  7.858653  7.839620  7.821870  7.805316
#> 12 26.268530 26.287805 26.353779 26.415304 26.472683 26.526194 26.576099
#> 13 29.600198 29.597709 29.589191 29.581246 29.573838 29.566928 29.560484
#> 14 28.347263 28.374587 28.468111 28.555328 28.636668 28.712525 28.783270
#> 15 24.759313 24.790765 24.898421 24.998818 25.092448 25.179767 25.261202
#> 16 18.750920 18.718990 18.609696 18.507772 18.412717 18.324069 18.241395
#> 17 19.235009 19.211928 19.132926 19.059252 18.990543 18.926464 18.866705
#> 18 27.777747 27.784210 27.806331 27.826960 27.846198 27.864140 27.880873
#> 19 26.610240 26.570765 26.435651 26.309649 26.192138 26.082548 25.980343
#> 20 30.181104 30.144066 30.017288 29.899059 29.788799 29.685969 29.590070
#> 21 22.965338 22.967874 22.976555 22.984651 22.992201 22.999243 23.005809
#>       0.7211    0.6725    0.6272    0.5849    0.5455    0.5087    0.4745
#> 1  23.083339 22.986111 22.895435 22.810870 22.745155 22.684051 22.627066
#> 2  21.862043 21.793125 21.728852 21.668911 21.683469 21.698337 21.712224
#> 3  25.611412 25.614636 25.617642 25.620446 25.624657 25.628616 25.632310
#> 4  18.673977 18.608700 18.547822 18.491047 18.195064 17.914366 17.652502
#> 5  19.804354 19.891593 19.972953 20.048829 19.943828 19.842624 19.748185
#> 6  21.242383 21.317405 21.387372 21.452623 21.700244 21.934777 22.153570
#> 7  22.594762 22.807629 23.006149 23.191289 23.601159 23.988061 24.348974
#> 8  19.404374 19.428624 19.451240 19.472332 19.702156 19.920501 20.124202
#> 9  15.266777 15.223654 15.183438 15.145933 14.989548 14.841367 14.703130
#> 10  8.833851  8.825952  8.818585  8.811715  8.909091  9.001871  9.088433
#> 11  7.789878  7.775481  7.762054  7.749531  7.869959  7.984771  8.091889
#> 12 26.622641 26.666046 26.706525 26.744276 26.776458 26.806442 26.834407
#> 13 29.554475 29.548870 29.543643 29.538769 29.406380 29.280480 29.163022
#> 14 28.849246 28.910777 28.968160 29.021676 29.011024 28.999982 28.989666
#> 15 25.337148 25.407976 25.474030 25.535632 25.232986 24.943942 24.674261
#> 16 18.164294 18.092388 18.025329 17.962790 17.672933 17.398161 17.141828
#> 17 18.810973 18.758997 18.710524 18.665318 18.384789 18.118600 17.870270
#> 18 27.896478 27.911031 27.924604 27.937262 27.886721 27.838413 27.793339
#> 19 25.885027 25.796134 25.713233 25.635919 25.910623 26.173342 26.418468
#> 20 29.500635 29.417227 29.339440 29.266896 29.436501 29.599126 29.750867
#> 21 23.011934 23.017645 23.022971 23.027939 23.117257 23.202169 23.281387
#>       0.4425    0.4127    0.3849    0.3589    0.3347    0.3122    0.2911
#> 1  22.573918 22.524356 22.516332 23.063160 23.063262 23.063327 23.056636
#> 2  21.725063 21.737149 21.744136 22.516428 22.516616 22.516736 22.513679
#> 3  25.635751 25.638963 25.639650 26.572531 26.572693 26.572798 26.554689
#> 4  17.408733 17.180951 17.123349 16.798493 16.798239 16.798076 16.769146
#> 5  19.660407 19.578251 19.551459 19.976111 19.976047 19.976005 19.939334
#> 6  22.357267 22.547583 22.594597 20.935558 20.935423 20.935337 20.994452
#> 7  24.685097 24.999029 25.071742 24.070150 24.070174 24.070190 24.121137
#> 8  20.313794 20.490985 20.537234 18.716867 18.716722 18.716629 18.783499
#> 9  14.574435 14.454190 14.424253 14.423730 14.423631 14.423567 14.406166
#> 10  9.168977  9.244275  9.264918  9.841327  9.841509  9.841627  9.846155
#> 11  8.191554  8.284734  8.310529  8.915079  8.915289  8.915424  8.923810
#> 12 26.860484 26.884806 26.888978 27.976279 27.976466 27.976586 27.953910
#> 13 29.053710 28.951537 28.924414 29.572084 29.572093 29.572099 29.540236
#> 14 28.980141 28.971164 28.965412 30.015563 30.015698 30.015785 29.984776
#> 15 24.423384 24.188792 24.121970 24.124851 24.124565 24.124381 24.068509
#> 16 16.903198 16.680227 16.624276 16.294635 16.294390 16.294232 16.267270
#> 17 17.639113 17.423102 17.367938 17.106133 17.105894 17.105741 17.076000
#> 18 27.751412 27.712204 27.700888 28.556656 28.556754 28.556818 28.530336
#> 19 26.646472 26.859706 26.921775 25.412835 25.412850 25.412859 25.496620
#> 20 29.891975 30.023976 30.063941 28.366282 28.366176 28.366108 28.436924
#> 21 23.355113 23.424022 23.442210 24.545249 24.545508 24.545674 24.536715
#>       0.2715    0.2532    0.2361    0.2202    0.2054    0.1915    0.1786
#> 1  23.022726 22.991006 22.978307 22.377303 22.186527 22.186099 22.185779
#> 2  22.494639 22.476767 22.469433 21.932019 21.761801 21.761392 21.761084
#> 3  26.466508 26.384087 26.351425 26.729602 26.850222 26.850302 26.850352
#> 4  16.640660 16.520792 16.473869 16.614954 16.658624 16.658687 16.658740
#> 5  19.770691 19.613246 19.551403 19.820314 19.905245 19.905302 19.905343
#> 6  21.272839 21.532868 21.635491 21.763618 21.804445 21.804814 21.805095
#> 7  24.358071 24.579325 24.666584 24.477186 24.417592 24.417709 24.417799
#> 8  19.098057 19.391863 19.507760 18.972560 18.802549 18.802564 18.802586
#> 9  14.327558 14.254197 14.225404 14.179187 14.163974 14.163914 14.163872
#> 10  9.862695  9.878055  9.883861  9.891238  9.894268  9.894212  9.894164
#> 11  8.957596  8.989046  9.001147  9.107718  9.142411  9.142414  9.142408
#> 12 27.844020 27.741319 27.700672 27.933274 28.007671 28.007648 28.007622
#> 13 29.391849 29.253279 29.198698 28.803259 28.677288 28.676936 28.676675
#> 14 29.837476 29.699866 29.645554 29.660365 29.665272 29.665118 29.664997
#> 15 23.815981 23.580307 23.487932 23.305411 23.245876 23.245691 23.245562
#> 16 16.147713 16.036180 15.992522 16.187082 16.247796 16.247892 16.247970
#> 17 16.943412 16.819709 16.771265 16.881152 16.914953 16.914991 16.915026
#> 18 28.404835 28.287597 28.241316 28.406604 28.459248 28.459203 28.459164
#> 19 25.885432 26.248496 26.391370 26.845509 26.990688 26.991243 26.991657
#> 20 28.768525 29.078221 29.200240 29.335701 29.379033 29.379408 29.379694
#> 21 24.488716 24.443774 24.425747 24.575943 24.624518 24.624463 24.624411
#>       0.1666    0.1554    0.1449    0.1351    0.1260    0.1175    0.1096
#> 1  22.185541 22.185366 22.185239 22.185146 21.589494 21.589591 21.589678
#> 2  21.760852 21.760680 21.760553 21.760460 21.439613 21.439792 21.439950
#> 3  26.850382 26.850400 26.850408 26.850411 26.441983 26.442048 26.442107
#> 4  16.658785 16.658822 16.658852 16.658876 17.494016 17.494052 17.494080
#> 5  19.905374 19.905396 19.905412 19.905423 19.317450 19.317716 19.317948
#> 6  21.805309 21.805470 21.805590 21.805679 20.975228 20.974826 20.974477
#> 7  24.417868 24.417921 24.417961 24.417991 25.169794 25.169952 25.170086
#> 8  18.802612 18.802637 18.802661 18.802682 19.242371 19.242190 19.242030
#> 9  14.163844 14.163824 14.163812 14.163803 14.433686 14.433722 14.433753
#> 10  9.894124  9.894090  9.894063  9.894041  9.679510  9.679525  9.679540
#> 11  9.142398  9.142385  9.142373  9.142361  9.051656  9.051589  9.051532
#> 12 28.007596 28.007571 28.007549 28.007530 28.377959 28.378156 28.378327
#> 13 28.676482 28.676340 28.676238 28.676164 29.611337 29.611140 29.610968
#> 14 29.664903 29.664831 29.664775 29.664733 30.405305 30.405606 30.405866
#> 15 23.245474 23.245415 23.245376 23.245351 21.745864 21.745350 21.744908
#> 16 16.248032 16.248081 16.248120 16.248150 16.237141 16.237255 16.237353
#> 17 16.915057 16.915083 16.915105 16.915123 17.856812 17.856910 17.856991
#> 18 28.459131 28.459103 28.459080 28.459062 28.548510 28.548624 28.548724
#> 19 26.991963 26.992188 26.992352 26.992470 27.532388 27.532108 27.531861
#> 20 29.379909 29.380071 29.380191 29.380280 28.485317 28.485381 28.485437
#> 21 24.624365 24.624325 24.624291 24.624263 24.164565 24.164465 24.164382
#>       0.1022    0.0953    0.0889    0.0829    0.0773    0.0721    0.0673
#> 1  21.589757 21.589828 21.589892 21.589949 21.590000 21.631737 21.631934
#> 2  21.440091 21.440214 21.440323 21.440419 21.440503 21.274141 21.274262
#> 3  26.442159 26.442206 26.442249 26.442286 26.442320 26.223513 26.223513
#> 4  17.494101 17.494117 17.494128 17.494136 17.494140 17.698004 17.698014
#> 5  19.318151 19.318328 19.318482 19.318615 19.318731 19.784050 19.783960
#> 6  20.974175 20.973913 20.973686 20.973490 20.973322 21.766077 21.765962
#> 7  25.170200 25.170295 25.170376 25.170444 25.170501 24.658140 24.658234
#> 8  19.241889 19.241765 19.241655 19.241559 19.241474 19.245314 19.245017
#> 9  14.433779 14.433801 14.433819 14.433834 14.433846 13.082237 13.082163
#> 10  9.679555  9.679569  9.679582  9.679594  9.679606 10.330385 10.330518
#> 11  9.051483  9.051443  9.051408  9.051379  9.051355  9.137285  9.137349
#> 12 28.378476 28.378605 28.378717 28.378813 28.378897 28.363241 28.363080
#> 13 29.610817 29.610686 29.610572 29.610473 29.610387 30.487815 30.487698
#> 14 30.406092 30.406287 30.406455 30.406599 30.406723 30.441676 30.441615
#> 15 21.744527 21.744201 21.743922 21.743683 21.743479 21.159731 21.160221
#> 16 16.237436 16.237508 16.237569 16.237620 16.237664 16.510297 16.510227
#> 17 17.857058 17.857113 17.857158 17.857195 17.857225 17.828515 17.828344
#> 18 28.548811 28.548887 28.548953 28.549011 28.549061 28.753219 28.753138
#> 19 27.531644 27.531452 27.531284 27.531137 27.531009 27.104926 27.104865
#> 20 28.485486 28.485528 28.485564 28.485596 28.485623 28.626920 28.627072
#> 21 24.164313 24.164255 24.164207 24.164167 24.164134 23.692777 23.692813
#>      0.0627    0.0585    0.0546    0.0509    0.0474    0.0442    0.0413
#> 1  21.63212 21.632304 21.632477 21.632642 21.632800 21.632950 21.633093
#> 2  21.27438 21.274492 21.274601 21.274707 21.274808 21.274906 21.275000
#> 3  26.22351 26.223517 26.223521 26.223526 26.223533 26.223540 26.223548
#> 4  17.69802 17.698020 17.698017 17.698010 17.698000 17.697987 17.697971
#> 5  19.78387 19.783779 19.783689 19.783598 19.783508 19.783419 19.783330
#> 6  21.76585 21.765746 21.765645 21.765548 21.765455 21.765367 21.765282
#> 7  24.65833 24.658415 24.658501 24.658584 24.658665 24.658743 24.658818
#> 8  19.24473 19.244465 19.244209 19.243965 19.243734 19.243514 19.243307
#> 9  13.08210 13.082049 13.082008 13.081976 13.081954 13.081940 13.081934
#> 10 10.33064 10.330752 10.330856 10.330951 10.331037 10.331115 10.331186
#> 11  9.13741  9.137469  9.137525  9.137578  9.137629  9.137677  9.137724
#> 12 28.36293 28.362778 28.362636 28.362500 28.362370 28.362245 28.362127
#> 13 30.48758 30.487466 30.487351 30.487238 30.487126 30.487015 30.486906
#> 14 30.44155 30.441493 30.441432 30.441371 30.441311 30.441251 30.441192
#> 15 21.16070 21.161162 21.161613 21.162050 21.162473 21.162883 21.163279
#> 16 16.51016 16.510085 16.510014 16.509944 16.509874 16.509805 16.509737
#> 17 17.82818 17.828020 17.827867 17.827718 17.827576 17.827438 17.827306
#> 18 28.75306 28.752983 28.752908 28.752834 28.752763 28.752693 28.752626
#> 19 27.10481 27.104763 27.104722 27.104687 27.104658 27.104635 27.104617
#> 20 28.62721 28.627348 28.627472 28.627587 28.627693 28.627792 28.627882
#> 21 23.69285 23.692894 23.692938 23.692985 23.693034 23.693084 23.693136
#>       0.0385    0.0359    0.0335    0.0312    0.0291    0.0272    0.0253
#> 1  21.633229 21.633358 21.633480 21.633596 21.633706 21.633809 21.633906
#> 2  21.275091 21.275178 21.275261 21.275341 21.275417 21.275490 21.275559
#> 3  26.223556 26.223566 26.223576 26.223587 26.223598 26.223610 26.223622
#> 4  17.697953 17.697932 17.697909 17.697884 17.697857 17.697829 17.697799
#> 5  19.783243 19.783156 19.783070 19.782986 19.782903 19.782821 19.782741
#> 6  21.765201 21.765125 21.765052 21.764982 21.764917 21.764855 21.764796
#> 7  24.658890 24.658959 24.659026 24.659091 24.659152 24.659211 24.659268
#> 8  19.243110 19.242925 19.242751 19.242586 19.242432 19.242288 19.242153
#> 9  13.081935 13.081943 13.081958 13.081979 13.082005 13.082036 13.082072
#> 10 10.331249 10.331305 10.331355 10.331398 10.331436 10.331468 10.331494
#> 11  9.137767  9.137809  9.137849  9.137886  9.137921  9.137955  9.137986
#> 12 28.362013 28.361906 28.361803 28.361706 28.361613 28.361526 28.361443
#> 13 30.486798 30.486693 30.486589 30.486488 30.486389 30.486292 30.486197
#> 14 30.441133 30.441076 30.441019 30.440962 30.440907 30.440853 30.440800
#> 15 21.163662 21.164032 21.164388 21.164731 21.165061 21.165378 21.165682
#> 16 16.509669 16.509603 16.509537 16.509472 16.509409 16.509347 16.509286
#> 17 17.827179 17.827058 17.826941 17.826829 17.826723 17.826621 17.826524
#> 18 28.752560 28.752496 28.752435 28.752375 28.752317 28.752261 28.752208
#> 19 27.104604 27.104595 27.104591 27.104591 27.104595 27.104602 27.104612
#> 20 28.627966 28.628042 28.628111 28.628174 28.628230 28.628281 28.628326
#> 21 23.693190 23.693244 23.693299 23.693356 23.693412 23.693469 23.693527
#>       0.0236    0.0220    0.0205    0.0192    0.0179    0.0167    0.0155
#> 1  21.633998 21.634084 21.634164 21.634240 21.634310 21.634376 21.634437
#> 2  21.275625 21.275688 21.275748 21.275805 21.275859 21.275910 21.275959
#> 3  26.223634 26.223647 26.223660 26.223673 26.223686 26.223699 26.223713
#> 4  17.697768 17.697736 17.697703 17.697669 17.697634 17.697599 17.697563
#> 5  19.782663 19.782586 19.782511 19.782438 19.782366 19.782296 19.782229
#> 6  21.764741 21.764688 21.764639 21.764593 21.764550 21.764509 21.764472
#> 7  24.659322 24.659374 24.659423 24.659470 24.659515 24.659557 24.659597
#> 8  19.242026 19.241909 19.241800 19.241699 19.241605 19.241519 19.241440
#> 9  13.082112 13.082156 13.082204 13.082254 13.082308 13.082364 13.082422
#> 10 10.331516 10.331533 10.331546 10.331554 10.331558 10.331559 10.331557
#> 11  9.138016  9.138043  9.138069  9.138094  9.138116  9.138137  9.138157
#> 12 28.361365 28.361291 28.361221 28.361156 28.361095 28.361037 28.360984
#> 13 30.486105 30.486015 30.485927 30.485842 30.485759 30.485679 30.485602
#> 14 30.440748 30.440698 30.440648 30.440600 30.440553 30.440507 30.440462
#> 15 21.165973 21.166253 21.166520 21.166775 21.167019 21.167251 21.167472
#> 16 16.509226 16.509168 16.509111 16.509056 16.509002 16.508950 16.508899
#> 17 17.826431 17.826343 17.826260 17.826180 17.826105 17.826034 17.825967
#> 18 28.752156 28.752106 28.752058 28.752011 28.751967 28.751925 28.751884
#> 19 27.104625 27.104641 27.104660 27.104681 27.104704 27.104729 27.104755
#> 20 28.628366 28.628401 28.628431 28.628456 28.628477 28.628494 28.628507
#> 21 23.693584 23.693641 23.693698 23.693755 23.693812 23.693868 23.693923
#>       0.0145    0.0135    0.0126    0.0118    0.0110    0.0102    0.0095
#> 1  21.634493 21.634545 21.634593 21.634637 21.634678 21.634715 21.634748
#> 2  21.276004 21.276047 21.276088 21.276126 21.276161 21.276194 21.276226
#> 3  26.223726 26.223740 26.223754 26.223767 26.223780 26.223794 26.223807
#> 4  17.697527 17.697490 17.697454 17.697418 17.697381 17.697345 17.697308
#> 5  19.782163 19.782099 19.782037 19.781977 19.781919 19.781862 19.781808
#> 6  21.764436 21.764404 21.764374 21.764346 21.764320 21.764296 21.764274
#> 7  24.659636 24.659672 24.659706 24.659738 24.659769 24.659797 24.659824
#> 8  19.241368 19.241302 19.241242 19.241188 19.241140 19.241097 19.241060
#> 9  13.082483 13.082545 13.082608 13.082673 13.082738 13.082804 13.082871
#> 10 10.331551 10.331543 10.331531 10.331518 10.331502 10.331484 10.331464
#> 11  9.138175  9.138192  9.138208  9.138222  9.138235  9.138247  9.138257
#> 12 28.360934 28.360887 28.360844 28.360804 28.360767 28.360733 28.360702
#> 13 30.485527 30.485454 30.485384 30.485316 30.485251 30.485188 30.485128
#> 14 30.440419 30.440377 30.440336 30.440297 30.440259 30.440222 30.440187
#> 15 21.167682 21.167882 21.168071 21.168251 21.168420 21.168580 21.168731
#> 16 16.508849 16.508801 16.508755 16.508710 16.508667 16.508625 16.508584
#> 17 17.825903 17.825844 17.825787 17.825735 17.825685 17.825639 17.825596
#> 18 28.751845 28.751808 28.751772 28.751738 28.751706 28.751675 28.751646
#> 19 27.104783 27.104813 27.104844 27.104876 27.104908 27.104942 27.104976
#> 20 28.628517 28.628523 28.628527 28.628527 28.628524 28.628519 28.628512
#> 21 23.693978 23.694032 23.694085 23.694138 23.694189 23.694240 23.694289
#>       0.0089    0.0083    0.0077   0.0072    0.0067    0.0063    0.0058
#> 1  21.634778 21.634805 21.634829 21.63485 21.634868 21.634884 21.634898
#> 2  21.276255 21.276282 21.276307 21.27633 21.276351 21.276371 21.276388
#> 3  26.223820 26.223833 26.223845 26.22386 26.223870 26.223882 26.223894
#> 4  17.697273 17.697237 17.697202 17.69717 17.697133 17.697099 17.697065
#> 5  19.781756 19.781705 19.781657 19.78161 19.781565 19.781522 19.781481
#> 6  21.764255 21.764237 21.764221 21.76421 21.764193 21.764182 21.764172
#> 7  24.659849 24.659873 24.659895 24.65991 24.659934 24.659952 24.659968
#> 8  19.241027 19.240998 19.240974 19.24095 19.240937 19.240925 19.240915
#> 9  13.082939 13.083006 13.083073 13.08314 13.083208 13.083274 13.083340
#> 10 10.331442 10.331418 10.331394 10.33137 10.331341 10.331313 10.331284
#> 11  9.138267  9.138276  9.138283  9.13829  9.138296  9.138301  9.138305
#> 12 28.360674 28.360648 28.360625 28.36060 28.360585 28.360569 28.360554
#> 13 30.485070 30.485015 30.484961 30.48491 30.484862 30.484815 30.484771
#> 14 30.440153 30.440120 30.440088 30.44006 30.440029 30.440001 30.439974
#> 15 21.168873 21.169007 21.169132 21.16925 21.169358 21.169460 21.169555
#> 16 16.508545 16.508508 16.508472 16.50844 16.508404 16.508373 16.508342
#> 17 17.825556 17.825519 17.825485 17.82545 17.825424 17.825397 17.825372
#> 18 28.751619 28.751592 28.751568 28.75154 28.751522 28.751502 28.751482
#> 19 27.105011 27.105046 27.105082 27.10512 27.105153 27.105189 27.105225
#> 20 28.628503 28.628491 28.628478 28.62846 28.628446 28.628428 28.628409
#> 21 23.694338 23.694385 23.694431 23.69448 23.694520 23.694562 23.694604
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 22.8 18.7 18.1 24.4 22.8 17.8 16.4 10.4 10.4 32.4 30.4 33.9 21.5
#> [16] 15.5 15.2 27.3 26.0 30.4 21.4
#> 
#> $X
#>               am       carb        cyl       disp        drat       gear
#>  [1,]  1.0488088  1.5892861  0.2744106 -0.2748944  0.38263377  0.4548588
#>  [2,]  1.0488088  1.5892861  0.2744106 -0.2748944  0.38263377  0.4548588
#>  [3,]  1.0488088 -1.1167956 -0.8781141 -0.7044046  0.29491206  0.4548588
#>  [4,] -0.9534626 -0.2147684  1.4269354  1.3770679 -0.93319198 -1.1371471
#>  [5,] -0.9534626 -1.1167956  0.2744106  0.2619934 -1.61742137 -1.1371471
#>  [6,] -0.9534626 -0.2147684 -0.8781141 -0.3847499  0.01420256  0.4548588
#>  [7,] -0.9534626 -0.2147684 -0.8781141 -0.4334828  0.41772246  0.4548588
#>  [8,] -0.9534626  1.5892861  0.2744106 -0.2121198  0.41772246  0.4548588
#>  [9,] -0.9534626  0.6872588  1.4269354  0.6815918 -1.07354672 -1.1371471
#> [10,] -0.9534626  1.5892861  1.4269354  2.3021668 -1.31916753 -1.1371471
#> [11,] -0.9534626  1.5892861  1.4269354  2.2030491 -1.19635713 -1.1371471
#> [12,]  1.0488088 -1.1167956 -0.8781141 -0.9464171  0.69843195  0.4548588
#> [13,]  1.0488088 -0.2147684 -0.8781141 -0.9711965  2.18970114  0.4548588
#> [14,]  1.0488088 -1.1167956 -0.8781141 -1.0091916  0.94405276  0.4548588
#> [15,] -0.9534626 -1.1167956 -0.8781141 -0.6044609  0.03174691 -1.1371471
#> [16,] -0.9534626 -0.2147684  1.4269354  1.0301558 -1.61742137 -1.1371471
#> [17,] -0.9534626 -0.2147684  1.4269354  0.9145185 -0.93319198 -1.1371471
#> [18,]  1.0488088 -1.1167956 -0.8781141 -0.9439391  0.69843195  0.4548588
#> [19,]  1.0488088 -0.2147684 -0.8781141 -0.6028089  1.31248397  2.0468647
#> [20,]  1.0488088 -0.2147684 -0.8781141 -0.8109562  0.15455731  2.0468647
#> [21,]  1.0488088 -0.2147684 -0.8781141 -0.5970270  0.75106498  0.4548588
#>                hp        qsec         vs         wt
#>  [1,] -0.12623260 -1.28573992 -1.1547005 -0.3550058
#>  [2,] -0.12623260 -0.92435604 -1.1547005 -0.1067289
#>  [3,] -0.49561814  0.10171604  0.8660254 -0.6470962
#>  [4,]  1.28612391 -0.92435604 -1.1547005  0.4433747
#>  [5,] -0.23487540  1.14069469  0.8660254  0.4628474
#>  [6,] -1.16920355  0.99872245  0.8660254  0.1999660
#>  [7,] -0.45216102  2.87017468  0.8660254  0.1610206
#>  [8,]  0.15623870  0.28886126  0.8660254  0.4433747
#>  [9,]  1.39476672 -0.67913127 -1.1547005  1.0567646
#> [10,]  1.93798076 -0.30484082 -1.1547005  2.2056535
#> [11,]  2.15526637 -0.40809336 -1.1547005  2.3750660
#> [12,] -1.08228931  0.65669842  0.8660254 -0.7639323
#> [13,] -1.38648917  0.04363649  0.8660254 -1.3335087
#> [14,] -1.10401787  0.93418962  0.8660254 -1.1193090
#> [15,] -0.40870390  1.00517574  0.8660254 -0.5059191
#> [16,]  0.74290987 -1.02115530 -1.1547005  0.5212655
#> [17,]  0.74290987 -0.74366410 -1.1547005  0.4385065
#> [18,] -1.08228931  0.28886126  0.8660254 -1.0219455
#> [19,] -0.53907527 -1.13086112 -1.1547005 -0.8223504
#> [20,] -0.06104691 -1.00179544  0.8660254 -1.4328194
#> [21,] -0.14796116  0.09526276  0.8660254 -0.1992242
#> attr(,"center")
#>  [1]   0.4761905   2.2380952   5.5238095 193.2809524   3.6819048   3.7142857
#>  [7] 115.8095238  18.4523810   0.5714286   2.9846190
#> attr(,"scale")
#>  [1]   0.4994328   1.1086140   1.7353207 121.0681375   0.5699843   0.6281384
#>  [7]  46.0223745   1.5495987   0.4948717   1.0270792
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 18
#> 
#> $lambda.min
#> [1] 1.786316
#> 
#> $null.dev
#> [1] 41.7381
#> 
#> $Bias
#> [1] 2.669449
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 11.70208