Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 35.67146 33.20054 31.25003 29.80474 28.78195 28.11084 26.18513 24.63804
#>   [9] 23.43437 22.51981 21.84779 21.37841 21.07750 20.91590 20.86871 20.83099
#>  [17] 20.31333 19.82452 19.25094 18.52601 17.88420 17.29651 16.98560 16.19992
#>  [25] 15.78797 15.29552 14.71219 14.53183 14.87042 17.79760 17.63520 17.48389
#>  [33] 17.52766 18.09393 17.70456 16.92363 16.32799 15.91796 15.89321 15.87021
#>  [41] 15.84885 15.82898 15.43576 15.80233 15.80209 15.80219 15.85213 16.33026
#>  [49] 16.66053 16.66063 16.38283 16.38278 17.52307 17.55398 17.66856 17.66848
#>  [57] 20.94285 20.94442 21.02794 21.13037 21.90169 21.74214 21.58628 21.75568
#>  [65] 21.58737 26.63924 27.37126 28.11525 28.16095 28.13550 29.54755 29.54798
#>  [73] 29.66093 29.66152 29.85729 30.10969 30.14199 30.23808 30.29261 30.29291
#>  [81] 30.29320 30.32909 30.35401 30.35440 30.35477 30.34346 30.21500 30.21450
#>  [89] 30.21406 30.21366 30.21332 30.21302 30.21277 30.21255 30.23960 30.24014
#>  [97] 30.38055 30.38131 30.38207 30.38283
#> 
#> $cvse
#>   [1] 13.512790 13.513567 13.506693 13.495083 13.476745 13.452528 12.382251
#>   [8] 11.401766 10.523238  9.741591  9.052918  8.454067  7.942200  7.514359
#>  [15]  7.167060  6.891420  6.566264  6.280692  6.008036  5.767499  5.550063
#>  [22]  5.386263  5.365864  5.370351  5.368884  5.369100  5.176089  4.946490
#>  [29]  4.830963  6.318244  6.359844  6.336177  6.466017  6.663867  6.558496
#>  [36]  6.415139  6.336792  6.367458  6.357421  6.348151  6.339585  6.331667
#>  [43]  6.072141  6.036722  6.036681  6.036636  6.029893  6.055256  6.047112
#>  [50]  6.047121  5.998215  5.998204  6.171820  6.167473  6.164930  6.164855
#>  [57]  8.897854  8.898923  8.895012  8.890601  8.892192  8.911603  8.953889
#>  [64]  8.947072  8.927374 13.750266 13.778434 13.828818 13.826758 13.824271
#>  [71] 13.694479 13.693967 13.687647 13.687140 13.693011 13.704016 13.701425
#>  [78] 13.690780 13.702198 13.701729 13.701265 13.696983 13.698043 13.697600
#>  [85] 13.697161 13.693251 13.699916 13.699518 13.699125 13.698736 13.698351
#>  [92] 13.697971 13.697595 13.697223 13.695371 13.694969 13.681067 13.680657
#>  [99] 13.680252 13.679851
#> 
#> $fold
#>  [1]  1  5  3  4  8  3  9  6  7  2  4 10  6  7  9  2  5 10  1  1  8
#> 
#> $lambda
#>   [1] 4.798664134 4.475250232 4.173633344 3.892344425 3.630013438 3.385362682
#>   [7] 3.157200567 2.944415816 2.745972045 2.560902719 2.388306447 2.227342586
#>  [13] 2.077227150 1.937228993 1.806666242 1.684902984 1.571346162 1.465442690
#>  [19] 1.366676757 1.274567318 1.188665747 1.108553654 1.033840848 0.964163435
#>  [25] 0.899182046 0.838580186 0.782062688 0.729354280 0.680198243 0.634355158
#>  [31] 0.591601743 0.551729766 0.514545026 0.479866413 0.447525023 0.417363334
#>  [37] 0.389234443 0.363001344 0.338536268 0.315720057 0.294441582 0.274597204
#>  [43] 0.256090272 0.238830645 0.222734259 0.207722715 0.193722899 0.180666624
#>  [49] 0.168490298 0.157134616 0.146544269 0.136667676 0.127456732 0.118866575
#>  [55] 0.110855365 0.103384085 0.096416343 0.089918205 0.083858019 0.078206269
#>  [61] 0.072935428 0.068019824 0.063435516 0.059160174 0.055172977 0.051454503
#>  [67] 0.047986641 0.044752502 0.041736333 0.038923444 0.036300134 0.033853627
#>  [73] 0.031572006 0.029444158 0.027459720 0.025609027 0.023883064 0.022273426
#>  [79] 0.020772272 0.019372290 0.018066662 0.016849030 0.015713462 0.014654427
#>  [85] 0.013666768 0.012745673 0.011886657 0.011085537 0.010338408 0.009641634
#>  [91] 0.008991820 0.008385802 0.007820627 0.007293543 0.006801982 0.006343552
#>  [97] 0.005916017 0.005517298 0.005145450 0.004798664
#> 
#> $fit
#> $beta
#>               4.7987     4.4753     4.1736     3.8923     3.6300    3.3854
#> (Intercept) 19.35714 21.1671737 22.8552146 24.4294871 25.8976589 27.266881
#> am           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> carb         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> cyl          0.00000 -0.2754395 -0.5323153 -0.7718785 -0.9952959 -1.203656
#> disp         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> drat         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> gear         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> hp           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> qsec         0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> vs           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> wt           0.00000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#>                3.1572    2.9444    2.7460    2.5609    2.3883    2.2273
#> (Intercept) 28.543822 29.734701 30.845320 31.881086 32.847045 33.747901
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.397973 -1.579194 -1.748201 -1.905817 -2.052811 -2.189898
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                2.0772    1.9372    1.8067    1.6849    1.5713    1.4654
#> (Intercept) 34.588043 35.371562 36.102275 36.783740 37.261404 37.261404
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.317746 -2.436977 -2.548172 -2.651873 -2.724561 -2.724561
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.3667    1.2746    1.1887     1.1086    1.0338    0.9642
#> (Intercept) 37.261404 37.261404 37.261404 37.6712334 38.203189 38.699206
#> am           0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> cyl         -2.724561 -2.724561 -2.724561 -2.5193114 -2.251689 -2.002285
#> disp         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.0000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000 -0.5283546 -1.216543 -1.857964
#>                0.8992    0.8386    0.7821    0.7294      0.6802      0.6344
#> (Intercept) 39.161905 39.425699 39.425441 39.425299 39.42880084 39.43812853
#> am           0.000000  0.000000  0.000000  0.000000  0.00000000  0.00000000
#> carb         0.000000  0.000000  0.000000  0.000000 -0.02508151 -0.08370862
#> cyl         -1.769455 -1.636401 -1.636263 -1.636188 -1.62685101 -1.60563753
#> disp         0.000000  0.000000  0.000000  0.000000  0.00000000  0.00000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.00000000  0.00000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.00000000  0.00000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.00000000  0.00000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.00000000  0.00000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.00000000  0.00000000
#> wt          -2.456654 -2.798596 -2.798791 -2.798898 -2.79541850 -2.78642262
#>                 0.5916     0.5517     0.5145     0.4799    0.4475    0.4174
#> (Intercept) 39.4465762 39.4548094 39.4625017 39.4696761 32.826447 32.826764
#> am           0.0000000  0.0000000  0.0000000  0.0000000  5.135619  5.135254
#> carb        -0.1383441 -0.1892424 -0.2367009 -0.2809604 -1.372696 -1.372602
#> cyl         -1.5857375 -1.5673928 -1.5502962 -1.5343523 -1.673566 -1.673640
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> wt          -2.7782241 -2.7703122 -2.7629231 -2.7560317  0.000000  0.000000
#>                0.3892    0.3630    0.3385    0.3157    0.2944    0.2746
#> (Intercept) 32.826928 32.827013 32.827057 32.827080 32.827092 32.827098
#> am           5.135064  5.134966  5.134915  5.134888  5.134875  5.134867
#> carb        -1.372552 -1.372527 -1.372513 -1.372507 -1.372503 -1.372501
#> cyl         -1.673678 -1.673698 -1.673708 -1.673714 -1.673716 -1.673718
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                0.2561    0.2388    0.2227    0.2077     0.1937     0.1807
#> (Intercept) 32.827102 32.827103 32.827104 32.827105 34.6040850 34.6044405
#> am           5.134864  5.134862  5.134861  5.134860  3.9963279  3.9958198
#> carb        -1.372500 -1.372500 -1.372499 -1.372499 -1.1735617 -1.1735051
#> cyl         -1.673719 -1.673719 -1.673719 -1.673719 -1.4789036 -1.4787445
#> disp         0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.0000000  0.0000000
#> wt           0.000000  0.000000  0.000000  0.000000 -0.9866297 -0.9870514
#>                 0.1685     0.1571     0.1465     0.1367     0.1275    0.1189
#> (Intercept) 34.6046915 34.6048686 34.6049936 39.4504584 39.4507371 37.245524
#> am           3.9954588  3.9952026  3.9950209  4.5009118  4.5013821  5.512878
#> carb        -1.1734653 -1.1734372 -1.1734175 -1.1573556 -1.1573779 -1.151986
#> cyl         -1.4786305 -1.4785490 -1.4784907 -1.6543936 -1.6545986 -1.110182
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> drat         0.0000000  0.0000000  0.0000000 -1.1453419 -1.1454569 -2.019535
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.000000
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  2.689807
#> wt          -0.9873522 -0.9875665 -0.9877189 -0.9604763 -0.9600614 -0.831364
#>                 0.1109     0.1034     0.0964     0.0899     0.0839     0.0782
#> (Intercept) 28.8457105 28.8441832 28.8427961 28.8415363 28.8403920 28.8393526
#> am           6.3028559  6.3036104  6.3043011  6.3049332  6.3055117  6.3060410
#> carb        -1.0243170 -1.0243183 -1.0243197 -1.0243212 -1.0243228 -1.0243243
#> cyl         -0.6963905 -0.6961223 -0.6958773 -0.6956536 -0.6954493 -0.6952629
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat        -2.4017968 -2.4024004 -2.4029531 -2.4034592 -2.4039224 -2.4043464
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.4276264  0.4276994  0.4277657  0.4278260  0.4278807  0.4279305
#> vs           2.7532721  2.7542416  2.7551294  2.7559424  2.7566865  2.7573676
#> wt          -1.1996508 -1.1996481 -1.1996447 -1.1996408 -1.1996366 -1.1996322
#>                 0.0729     0.0680     0.0634     0.0592     0.0552     0.0515
#> (Intercept) 28.8384083 28.8375504 28.8367709 28.8360626 28.8354189 28.8348339
#> am           6.3065252  6.3069680  6.3073730  6.3077432  6.3080816  6.3083910
#> carb        -1.0243259 -1.0243274 -1.0243290 -1.0243304 -1.0243319 -1.0243332
#> cyl         -0.6950927 -0.6949373 -0.6947955 -0.6946661 -0.6945480 -0.6944402
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat        -2.4047344 -2.4050893 -2.4054139 -2.4057108 -2.4059823 -2.4062305
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.4279757  0.4280168  0.4280542  0.4280882  0.4281191  0.4281472
#> vs           2.7579908  2.7585610  2.7590825  2.7595594  2.7599955  2.7603942
#> wt          -1.1996276 -1.1996230 -1.1996185 -1.1996139 -1.1996095 -1.1996051
#>                 0.0480     0.0448     0.0417     0.0389     0.0363     0.0339
#> (Intercept) 28.8343023 28.8338191 28.8333798 28.8329805 27.6046334 27.6032362
#> am           6.3086737  6.3089320  6.3091680  6.3093836  6.1156035  6.1164173
#> carb        -1.0243345 -1.0243358 -1.0243370 -1.0243381 -1.1001384 -1.1002461
#> cyl         -0.6943419 -0.6942522 -0.6941704 -0.6940957 -0.6773180 -0.6771178
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat        -2.4064574 -2.4066647 -2.4068542 -2.4070273 -2.5418758 -2.5429122
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.3567421  0.3570751
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.4281727  0.4281959  0.4282170  0.4282362  0.4582690  0.4583734
#> vs           2.7607586  2.7610917  2.7613961  2.7616742  2.5686096  2.5693388
#> wt          -1.1996010 -1.1995969 -1.1995931 -1.1995894 -1.1667724 -1.1666388
#>                   0.0316      0.0294      0.0275       0.0256       0.0239
#> (Intercept) 27.523852555 27.52261571 27.52157051 27.520696571 27.519975187
#> am           6.127156942  6.12808359  6.12893901  6.129727177  6.130452006
#> carb        -1.043018990 -1.04298707 -1.04295829 -1.042932365 -1.042909036
#> cyl         -0.569110660 -0.56886484 -0.56864593 -0.568451557 -0.568279551
#> disp         0.000000000  0.00000000  0.00000000  0.000000000  0.000000000
#> drat        -2.507211724 -2.50776264 -2.50826545 -2.508723327 -2.509139348
#> gear         0.372042752  0.37179862  0.37156590  0.371344626  0.371134753
#> hp          -0.003685965 -0.00368422 -0.00368255 -0.003680959 -0.003679449
#> qsec         0.426512901  0.42660512  0.42668706  0.426759613  0.426823592
#> vs           2.746762465  2.74781127  2.74877215  2.749650878  2.750452981
#> wt          -1.133835760 -1.13390441 -1.13396348 -1.134013938 -1.134056700
#>                   0.0223       0.0208       0.0194       0.0181       0.0168
#> (Intercept) 27.519389181 27.518922846 27.518561847 27.518293141 27.518104886
#> am           6.131117309  6.131726795  6.132284054  6.132792549  6.133255609
#> carb        -1.042888067 -1.042869243 -1.042852365 -1.042837252 -1.042823738
#> cyl         -0.568127874 -0.567994634 -0.567878082 -0.567776599 -0.567688690
#> disp         0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> drat        -2.509516455 -2.509857469 -2.510165077 -2.510441837 -2.510690169
#> gear         0.370936157  0.370748658  0.370572024  0.370405979  0.370250212
#> hp          -0.003678018 -0.003676669 -0.003675399 -0.003674207 -0.003673092
#> qsec         0.426879768  0.426928861  0.426971538  0.427008422  0.427040088
#> vs           2.751183752  2.751848240  2.752451252  2.752997351  2.753490856
#> wt          -1.134092589 -1.134122365 -1.134146722 -1.134166293 -1.134181654
#>                   0.0157       0.0147       0.0137       0.0127       0.0119
#> (Intercept) 27.517986365 27.517927903 27.517920791 27.517957213 27.518030181
#> am           6.133676418  6.134058018  6.134403304  6.134715024  6.134995776
#> carb        -1.042811671 -1.042800915 -1.042791342 -1.042782837 -1.042775296
#> cyl         -0.567612980 -0.567548203 -0.567493199 -0.567446907 -0.567408354
#> disp         0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> drat        -2.510912364 -2.511110577 -2.511286836 -2.511443039 -2.511580965
#> gear         0.370104383  0.369968128  0.369841066  0.369722806  0.369612946
#> hp          -0.003672051 -0.003671081 -0.003670181 -0.003669347 -0.003668575
#> qsec         0.427067069  0.427089853  0.427108894  0.427124603  0.427137361
#> vs           2.753935850  2.754336180  2.754695462  2.755017092  2.755304248
#> wt          -1.134193328 -1.134201788 -1.134207463 -1.134210738 -1.134211958
#>                   0.0111        0.0103       0.0096       0.0090       0.0084
#> (Intercept) 27.518133460  2.752154e+01 27.854144742 27.854126867 27.854093538
#> am           6.135248017  6.135234e+00  6.131417140  6.131327635  6.131245796
#> carb        -1.042768622 -1.042638e+00 -1.123443072 -1.123649427 -1.123846030
#> cyl         -0.567376657 -5.671943e-01 -0.571670454 -0.571587966 -0.571506957
#> disp         0.000000000 -8.266605e-06 -0.001948678 -0.001953120 -0.001957356
#> drat        -2.511702270 -2.512756e+00 -2.520641535 -2.520554162 -2.520473045
#> gear         0.369511081  3.692972e-01  0.402068770  0.402273354  0.402468647
#> hp          -0.003667864 -3.662993e-03 -0.001946066 -0.001944229 -0.001942490
#> qsec         0.427147512  4.271672e-01  0.396381145  0.396279223  0.396182404
#> vs           2.755559903  2.755995e+00  2.704205144  2.704242605  2.704284848
#> wt          -1.134211434 -1.134132e+00 -0.946741012 -0.946233984 -0.945750656
#>                   0.0078       0.0073       0.0068       0.0063       0.0059
#> (Intercept) 27.854046434 27.853987112 27.853917007 27.853837445 27.853749648
#> am           6.131171148  6.131103238  6.131041634  6.130985924  6.130935715
#> carb        -1.124033295 -1.124211624 -1.124381405 -1.124543010 -1.124696798
#> cyl         -0.571427501 -0.571349660 -0.571273489 -0.571199030 -0.571126320
#> disp        -0.001961394 -0.001965242 -0.001968908 -0.001972401 -0.001975728
#> drat        -2.520397816 -2.520328126 -2.520263641 -2.520204041 -2.520149024
#> gear         0.402654997  0.402832746  0.403002229  0.403163771  0.403317689
#> hp          -0.001940846 -0.001939292 -0.001937822 -0.001936431 -0.001935117
#> qsec         0.396090456  0.396003151  0.395920274  0.395841618  0.395766984
#> vs           2.704331339  2.704381580  2.704435103  2.704491471  2.704550281
#> wt          -0.945290026 -0.944851126 -0.944433025 -0.944034823 -0.943655657
#>                   0.0055       0.0051       0.0048
#> (Intercept) 27.853654737 27.853553743 27.853447612
#> am           6.130890635  6.130850329  6.130814461
#> carb        -1.124843116 -1.124982296 -1.125114657
#> cyl         -0.571055389 -0.570986255 -0.570918936
#> disp        -0.001978896 -0.001981912 -0.001984783
#> drat        -2.520098303 -2.520051604 -2.520008669
#> gear         0.403464293  0.403603882  0.403736750
#> hp          -0.001933875 -0.001932701 -0.001931591
#> qsec         0.395696179  0.395629023  0.395565339
#> vs           2.704611155  2.704673744  2.704737725
#> wt          -0.943294691 -0.942951126 -0.942624190
#> 
#> $iter
#>   [1]  0  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  1  1  1  1 28 29 28 28
#>  [26] 12  1  1  6  6  5  5  5  5 38  1  1  1  1  1  1  1  1  1  1  1 29  1  1  1
#>  [51]  1 43  1 57 69  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 36  1 63  1  1
#>  [76]  1  1  1  1  1  1  1  1  1  1  1  1  1  2 77  1  1  1  1  1  1  1  1  1  1
#> 
#> $lambda
#>   [1] 4.798664134 4.475250232 4.173633344 3.892344425 3.630013438 3.385362682
#>   [7] 3.157200567 2.944415816 2.745972045 2.560902719 2.388306447 2.227342586
#>  [13] 2.077227150 1.937228993 1.806666242 1.684902984 1.571346162 1.465442690
#>  [19] 1.366676757 1.274567318 1.188665747 1.108553654 1.033840848 0.964163435
#>  [25] 0.899182046 0.838580186 0.782062688 0.729354280 0.680198243 0.634355158
#>  [31] 0.591601743 0.551729766 0.514545026 0.479866413 0.447525023 0.417363334
#>  [37] 0.389234443 0.363001344 0.338536268 0.315720057 0.294441582 0.274597204
#>  [43] 0.256090272 0.238830645 0.222734259 0.207722715 0.193722899 0.180666624
#>  [49] 0.168490298 0.157134616 0.146544269 0.136667676 0.127456732 0.118866575
#>  [55] 0.110855365 0.103384085 0.096416343 0.089918205 0.083858019 0.078206269
#>  [61] 0.072935428 0.068019824 0.063435516 0.059160174 0.055172977 0.051454503
#>  [67] 0.047986641 0.044752502 0.041736333 0.038923444 0.036300134 0.033853627
#>  [73] 0.031572006 0.029444158 0.027459720 0.025609027 0.023883064 0.022273426
#>  [79] 0.020772272 0.019372290 0.018066662 0.016849030 0.015713462 0.014654427
#>  [85] 0.013666768 0.012745673 0.011886657 0.011085537 0.010338408 0.009641634
#>  [91] 0.008991820 0.008385802 0.007820627 0.007293543 0.006801982 0.006343552
#>  [97] 0.005916017 0.005517298 0.005145450 0.004798664
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 21
#> 
#> $loss
#>   [1] 665.65143 572.82047 495.15357 430.46850 376.88124 332.76615 296.72152
#>   [8] 267.53948 244.18009 225.74868 211.47631 200.70276 192.86178 187.46833
#>  [15] 184.10750 182.42489 182.08070 182.08070 182.08070 182.08070 182.08070
#>  [22] 160.42053 138.97789 125.88879 119.67537 118.72748 118.72748 118.72748
#>  [29] 117.52211 114.93629 112.81867 111.09947 109.71684 108.61904  85.69668
#>  [36]  85.69667  85.69667  85.69667  85.69667  85.69667  85.69667  85.69667
#>  [43]  85.69667  85.69667  85.69667  85.69667  81.76412  81.76411  81.76411
#>  [50]  81.76411  81.76411  80.29430  80.29430  76.99872  75.02183  75.02182
#>  [57]  75.02181  75.02180  75.02179  75.02178  75.02178  75.02177  75.02177
#>  [64]  75.02176  75.02176  75.02176  75.02176  75.02176  75.02175  75.02175
#>  [71]  74.80663  74.80661  74.62770  74.62769  74.62767  74.62766  74.62765
#>  [78]  74.62765  74.62764  74.62764  74.62763  74.62763  74.62763  74.62762
#>  [85]  74.62762  74.62762  74.62762  74.62762  74.62718  74.57270  74.57269
#>  [92]  74.57267  74.57266  74.57265  74.57264  74.57263  74.57262  74.57261
#>  [99]  74.57261  74.57260
#> 
#> $linear.predictors
#>      4.7987   4.4753   4.1736   3.8923   3.6300   3.3854   3.1572   2.9444
#> 1  19.35714 19.51454 19.66132 19.79822 19.92588 20.04495 20.15598 20.25954
#> 2  19.35714 18.96366 18.59669 18.25446 17.93529 17.63763 17.36004 17.10115
#> 3  19.35714 18.96366 18.59669 18.25446 17.93529 17.63763 17.36004 17.10115
#> 4  19.35714 20.06542 20.72595 21.34197 21.91648 22.45226 22.95193 23.41793
#> 5  19.35714 20.06542 20.72595 21.34197 21.91648 22.45226 22.95193 23.41793
#> 6  19.35714 19.51454 19.66132 19.79822 19.92588 20.04495 20.15598 20.25954
#> 7  19.35714 18.96366 18.59669 18.25446 17.93529 17.63763 17.36004 17.10115
#> 8  19.35714 18.96366 18.59669 18.25446 17.93529 17.63763 17.36004 17.10115
#> 9  19.35714 18.96366 18.59669 18.25446 17.93529 17.63763 17.36004 17.10115
#> 10 19.35714 18.96366 18.59669 18.25446 17.93529 17.63763 17.36004 17.10115
#> 11 19.35714 18.96366 18.59669 18.25446 17.93529 17.63763 17.36004 17.10115
#> 12 19.35714 20.06542 20.72595 21.34197 21.91648 22.45226 22.95193 23.41793
#> 13 19.35714 20.06542 20.72595 21.34197 21.91648 22.45226 22.95193 23.41793
#> 14 19.35714 18.96366 18.59669 18.25446 17.93529 17.63763 17.36004 17.10115
#> 15 19.35714 18.96366 18.59669 18.25446 17.93529 17.63763 17.36004 17.10115
#> 16 19.35714 18.96366 18.59669 18.25446 17.93529 17.63763 17.36004 17.10115
#> 17 19.35714 20.06542 20.72595 21.34197 21.91648 22.45226 22.95193 23.41793
#> 18 19.35714 20.06542 20.72595 21.34197 21.91648 22.45226 22.95193 23.41793
#> 19 19.35714 18.96366 18.59669 18.25446 17.93529 17.63763 17.36004 17.10115
#> 20 19.35714 19.51454 19.66132 19.79822 19.92588 20.04495 20.15598 20.25954
#> 21 19.35714 18.96366 18.59669 18.25446 17.93529 17.63763 17.36004 17.10115
#>      2.7460   2.5609   2.3883   2.2273   2.0772   1.9372   1.8067   1.6849
#> 1  20.35611 20.44618 20.53018 20.60851 20.68157 20.74970 20.81324 20.87250
#> 2  16.85971 16.63455 16.42456 16.22872 16.04608 15.87575 15.71690 15.56875
#> 3  16.85971 16.63455 16.42456 16.22872 16.04608 15.87575 15.71690 15.56875
#> 4  23.85252 24.25782 24.63580 24.98831 25.31706 25.62365 25.90959 26.17625
#> 5  23.85252 24.25782 24.63580 24.98831 25.31706 25.62365 25.90959 26.17625
#> 6  20.35611 20.44618 20.53018 20.60851 20.68157 20.74970 20.81324 20.87250
#> 7  16.85971 16.63455 16.42456 16.22872 16.04608 15.87575 15.71690 15.56875
#> 8  16.85971 16.63455 16.42456 16.22872 16.04608 15.87575 15.71690 15.56875
#> 9  16.85971 16.63455 16.42456 16.22872 16.04608 15.87575 15.71690 15.56875
#> 10 16.85971 16.63455 16.42456 16.22872 16.04608 15.87575 15.71690 15.56875
#> 11 16.85971 16.63455 16.42456 16.22872 16.04608 15.87575 15.71690 15.56875
#> 12 23.85252 24.25782 24.63580 24.98831 25.31706 25.62365 25.90959 26.17625
#> 13 23.85252 24.25782 24.63580 24.98831 25.31706 25.62365 25.90959 26.17625
#> 14 16.85971 16.63455 16.42456 16.22872 16.04608 15.87575 15.71690 15.56875
#> 15 16.85971 16.63455 16.42456 16.22872 16.04608 15.87575 15.71690 15.56875
#> 16 16.85971 16.63455 16.42456 16.22872 16.04608 15.87575 15.71690 15.56875
#> 17 23.85252 24.25782 24.63580 24.98831 25.31706 25.62365 25.90959 26.17625
#> 18 23.85252 24.25782 24.63580 24.98831 25.31706 25.62365 25.90959 26.17625
#> 19 16.85971 16.63455 16.42456 16.22872 16.04608 15.87575 15.71690 15.56875
#> 20 20.35611 20.44618 20.53018 20.60851 20.68157 20.74970 20.81324 20.87250
#> 21 16.85971 16.63455 16.42456 16.22872 16.04608 15.87575 15.71690 15.56875
#>      1.5713   1.4654   1.3667   1.2746   1.1887   1.1086   1.0338   0.9642
#> 1  20.91404 20.91404 20.91404 20.91404 20.91404 21.03635 21.19550 21.34385
#> 2  15.46491 15.46491 15.46491 15.46491 15.46491 15.69920 16.00477 16.28953
#> 3  15.46491 15.46491 15.46491 15.46491 15.46491 15.63052 15.84662 16.04799
#> 4  26.36316 26.36316 26.36316 26.36316 26.36316 25.90854 25.31566 24.76316
#> 5  26.36316 26.36316 26.36316 26.36316 26.36316 25.92967 25.36432 24.83748
#> 6  20.91404 20.91404 20.91404 20.91404 20.91404 20.73783 20.50815 20.29410
#> 7  15.46491 15.46491 15.46491 15.46491 15.46491 15.36634 15.23835 15.11901
#> 8  15.46491 15.46491 15.46491 15.46491 15.46491 15.54598 15.65197 15.75072
#> 9  15.46491 15.46491 15.46491 15.46491 15.46491 15.51956 15.59115 15.65782
#> 10 15.46491 15.46491 15.46491 15.46491 15.46491 14.74288 13.80283 12.92661
#> 11 15.46491 15.46491 15.46491 15.46491 15.46491 14.69269 13.68726 12.75011
#> 12 26.36316 26.36316 26.36316 26.36316 26.36316 26.62446 26.96408 27.28070
#> 13 26.36316 26.36316 26.36316 26.36316 26.36316 26.29159 26.19766 26.11018
#> 14 15.46491 15.46491 15.46491 15.46491 15.46491 15.65693 15.90745 16.14089
#> 15 15.46491 15.46491 15.46491 15.46491 15.46491 15.70184 16.01085 16.29882
#> 16 15.46491 15.46491 15.46491 15.46491 15.46491 15.48522 15.51207 15.53705
#> 17 26.36316 26.36316 26.36316 26.36316 26.36316 26.57162 26.84242 27.09490
#> 18 26.36316 26.36316 26.36316 26.36316 26.36316 26.79459 27.35580 27.87897
#> 19 15.46491 15.46491 15.46491 15.46491 15.46491 15.84186 16.33324 16.79118
#> 20 20.91404 20.91404 20.91404 20.91404 20.91404 21.09182 21.32323 21.53893
#> 21 15.46491 15.46491 15.46491 15.46491 15.46491 15.63052 15.84662 16.04799
#>      0.8992   0.8386   0.7821   0.7294   0.6802   0.6344   0.5916   0.5517
#> 1  21.48229 21.56133 21.56134 21.56134 21.53054 21.45850 21.39138 21.32884
#> 2  16.55537 16.70732 16.70749 16.70759 16.74759 16.84032 16.92690 17.00731
#> 3  16.23601 16.34350 16.34365 16.34373 16.33402 16.31067 16.28904 16.26868
#> 4  24.24736 23.95257 23.95224 23.95206 23.95385 23.95947 23.96440 23.96946
#> 5  24.34563 24.06451 24.06419 24.06402 24.06567 24.07093 24.07553 24.08027
#> 6  20.09428 19.98012 19.98002 19.97996 19.95113 19.88418 19.82168 19.76361
#> 7  15.00768 14.94420 14.94425 14.94428 14.96139 15.00116 15.03827 15.07277
#> 8  15.84294 15.89572 15.89584 15.89591 15.91184 15.94855 15.98287 16.01468
#> 9  15.72011 15.75579 15.75590 15.75596 15.77207 15.80922 15.84396 15.87616
#> 10 12.10883 11.64186 11.64168 11.64158 11.63772 11.62948 11.62162 11.61456
#> 11 11.87545 11.37599 11.37579 11.37569 11.37215 11.36476 11.35769 11.35138
#> 12 27.57612 27.74467 27.74461 27.74457 27.76672 27.81878 27.86724 27.91247
#> 13 26.02843 25.98155 25.98137 25.98126 26.00561 26.06334 26.11696 26.16718
#> 14 16.35884 16.48343 16.48359 16.48368 16.52396 16.61740 16.70464 16.78568
#> 15 16.56766 16.72131 16.72149 16.72158 16.76157 16.85425 16.94079 17.02116
#> 16 15.56043 15.57389 15.57398 15.57403 15.61545 15.71182 15.80172 15.88533
#> 17 27.33046 27.46481 27.46473 27.46468 27.48718 27.54014 27.58942 27.63544
#> 18 28.36717 28.64582 28.64582 28.64582 28.64177 28.63230 28.62349 28.61527
#> 19 17.21867 17.46294 17.46317 17.46329 17.45219 17.42523 17.40033 17.37681
#> 20 21.74024 21.85518 21.85521 21.85523 21.77390 21.58366 21.40641 21.24123
#> 21 16.23601 16.34350 16.34365 16.34373 16.23370 15.97583 15.73566 15.51171
#>      0.5145   0.4799   0.4475   0.4174   0.3892   0.3630   0.3385   0.3157
#> 1  21.27052 21.21613 22.42988 22.42977 22.42971 22.42968 22.42967 22.42966
#> 2  17.08227 17.15219 16.69252 16.69244 16.69240 16.69238 16.69236 16.69236
#> 3  16.24969 16.23198 13.94713 13.94724 13.94729 13.94732 13.94734 13.94734
#> 4  23.97419 23.97860 23.38679 23.38700 23.38711 23.38717 23.38720 23.38721
#> 5  24.08471 24.08885 23.38679 23.38700 23.38711 23.38717 23.38720 23.38721
#> 6  19.70947 19.65897 17.29426 17.29452 17.29465 17.29472 17.29475 17.29477
#> 7  15.10493 15.13493 15.31983 15.31984 15.31985 15.31985 15.31985 15.31985
#> 8  16.04433 16.07198 15.31983 15.31984 15.31985 15.31985 15.31985 15.31985
#> 9  15.90618 15.93418 15.31983 15.31984 15.31985 15.31985 15.31985 15.31985
#> 10 11.60798 11.60185 13.94713 13.94724 13.94729 13.94732 13.94734 13.94734
#> 11 11.34550 11.34003 13.94713 13.94724 13.94729 13.94732 13.94734 13.94734
#> 12 27.95465 27.99399 29.89510 29.89486 29.89473 29.89466 29.89463 29.89461
#> 13 26.21401 26.25769 24.75949 24.75960 24.75966 24.75969 24.75971 24.75972
#> 14 16.86124 16.93170 16.69252 16.69244 16.69240 16.69238 16.69236 16.69236
#> 15 17.09609 17.16597 16.69252 16.69244 16.69240 16.69238 16.69236 16.69236
#> 16 15.96329 16.03599 16.69252 16.69244 16.69240 16.69238 16.69236 16.69236
#> 17 27.67836 27.71838 29.89510 29.89486 29.89473 29.89466 29.89463 29.89461
#> 18 28.60761 28.60047 28.52241 28.52225 28.52217 28.52213 28.52211 28.52210
#> 19 17.35486 17.33439 19.08275 19.08249 19.08236 19.08229 19.08225 19.08223
#> 20 21.08722 20.94359 19.68449 19.68457 19.68461 19.68463 19.68464 19.68465
#> 21 15.30289 15.10814 13.59196 13.59209 13.59215 13.59218 13.59220 13.59221
#>      0.2944   0.2746   0.2561   0.2388   0.2227   0.2077   0.1937   0.1807
#> 1  22.42966 22.42965 22.42965 22.42965 22.42965 22.42965 22.19618 22.19600
#> 2  16.69235 16.69235 16.69235 16.69235 16.69235 16.69235 17.03173 17.03202
#> 3  13.94735 13.94735 13.94735 13.94735 13.94735 13.94735 14.55634 14.55669
#> 4  23.38722 23.38722 23.38723 23.38723 23.38723 23.38723 23.19400 23.19376
#> 5  23.38722 23.38722 23.38723 23.38723 23.38723 23.38723 23.23346 23.23324
#> 6  17.29478 17.29479 17.29479 17.29479 17.29479 17.29479 17.64241 17.64250
#> 7  15.31985 15.31985 15.31985 15.31985 15.31985 15.31985 15.23659 15.23667
#> 8  15.31985 15.31985 15.31985 15.31985 15.31985 15.31985 15.57204 15.57227
#> 9  15.31985 15.31985 15.31985 15.31985 15.31985 15.31985 15.52271 15.52291
#> 10 13.94735 13.94735 13.94735 13.94735 13.94735 13.94735 12.89880 12.89844
#> 11 13.94735 13.94735 13.94735 13.94735 13.94735 13.94735 12.80507 12.80467
#> 12 29.89460 29.89459 29.89459 29.89459 29.89459 29.89459 29.70077 29.70054
#> 13 24.75972 24.75973 24.75973 24.75973 24.75973 24.75973 25.08287 25.08288
#> 14 16.69235 16.69235 16.69235 16.69235 16.69235 16.69235 16.95280 16.95305
#> 15 16.69235 16.69235 16.69235 16.69235 16.69235 16.69235 17.03666 17.03695
#> 16 16.69235 16.69235 16.69235 16.69235 16.69235 16.69235 16.63214 16.63226
#> 17 29.89460 29.89459 29.89459 29.89459 29.89459 29.89459 29.60211 29.60183
#> 18 28.52210 28.52209 28.52209 28.52209 28.52209 28.52209 28.84490 28.84486
#> 19 19.08222 19.08222 19.08222 19.08221 19.08221 19.08221 18.94732 18.94733
#> 20 19.68465 19.68465 19.68465 19.68465 19.68465 19.68465 19.95266 19.95263
#> 21 13.59221 13.59221 13.59222 13.59222 13.59222 13.59222 13.85842 13.85849
#>      0.1685   0.1571   0.1465   0.1367   0.1275   0.1189   0.1109   0.1034
#> 1  22.19587 22.19577 22.19571 22.16738 22.16756 21.22301 21.33515 21.33488
#> 2  17.03223 17.03237 17.03248 16.98873 16.98839 16.83867 16.81170 16.81166
#> 3  14.55694 14.55712 14.55724 14.48044 14.48010 14.30545 13.95840 13.95824
#> 4  23.19359 23.19346 23.19337 23.22794 23.22826 23.08650 22.62780 22.62755
#> 5  23.23308 23.23296 23.23288 23.00293 23.00320 22.65526 23.36349 23.36331
#> 6  17.64256 17.64260 17.64263 17.10090 17.10083 17.88983 17.86367 17.86374
#> 7  15.23673 15.23677 15.23680 15.31790 15.31781 15.32449 15.38624 15.38628
#> 8  15.57243 15.57254 15.57262 15.64447 15.64423 15.60715 15.87965 15.87970
#> 9  15.52306 15.52316 15.52324 15.59644 15.59623 15.56558 15.99072 15.99080
#> 10 12.89819 12.89800 12.89787 13.18753 13.18793 13.47423 13.53061 13.53078
#> 11 12.80439 12.80419 12.80404 12.75269 12.75308 12.78939 12.45663 12.45658
#> 12 29.70037 29.70025 29.70017 29.58062 29.58081 29.80750 30.26478 30.26496
#> 13 25.08288 25.08288 25.08289 25.07019 25.07022 24.82103 24.50212 24.50187
#> 14 16.95324 16.95337 16.95346 17.35858 17.35832 17.55978 17.58828 17.58847
#> 15 17.03716 17.03731 17.03742 16.99354 16.99319 16.84283 16.93743 16.93742
#> 16 16.63235 16.63241 16.63245 16.67991 16.67975 16.64334 16.50679 16.50680
#> 17 29.60164 29.60150 29.60140 29.64492 29.64516 30.00710 30.05344 30.05363
#> 18 28.84483 28.84481 28.84480 29.24794 29.24802 29.83201 29.42468 29.42491
#> 19 18.94734 18.94734 18.94735 18.20875 18.20860 18.11114 17.74228 17.74217
#> 20 19.95261 19.95260 19.95258 20.27422 20.27434 19.57180 19.43499 19.43478
#> 21 13.85854 13.85857 13.85859 13.97397 13.97397 14.54393 14.84114 14.84144
#>      0.0964   0.0899   0.0839   0.0782   0.0729   0.0680   0.0634   0.0592
#> 1  21.33463 21.33440 21.33419 21.33400 21.33383 21.33367 21.33352 21.33339
#> 2  16.81163 16.81160 16.81158 16.81155 16.81153 16.81151 16.81149 16.81147
#> 3  13.95810 13.95796 13.95784 13.95773 13.95763 13.95754 13.95745 13.95738
#> 4  22.62733 22.62712 22.62694 22.62677 22.62661 22.62647 22.62634 22.62622
#> 5  23.36316 23.36301 23.36287 23.36275 23.36264 23.36253 23.36243 23.36235
#> 6  17.86380 17.86386 17.86391 17.86396 17.86400 17.86404 17.86408 17.86411
#> 7  15.38632 15.38636 15.38639 15.38642 15.38645 15.38647 15.38650 15.38652
#> 8  15.87975 15.87980 15.87984 15.87988 15.87992 15.87995 15.87998 15.88000
#> 9  15.99088 15.99095 15.99101 15.99107 15.99113 15.99117 15.99122 15.99126
#> 10 13.53094 13.53108 13.53122 13.53134 13.53145 13.53155 13.53165 13.53173
#> 11 12.45654 12.45650 12.45646 12.45643 12.45640 12.45637 12.45635 12.45633
#> 12 30.26513 30.26527 30.26541 30.26554 30.26565 30.26575 30.26585 30.26593
#> 13 24.50164 24.50143 24.50123 24.50106 24.50090 24.50075 24.50062 24.50049
#> 14 17.58865 17.58881 17.58895 17.58909 17.58921 17.58932 17.58942 17.58952
#> 15 16.93740 16.93739 16.93738 16.93737 16.93736 16.93735 16.93734 16.93734
#> 16 16.50681 16.50682 16.50683 16.50684 16.50685 16.50686 16.50687 16.50687
#> 17 30.05381 30.05397 30.05412 30.05425 30.05437 30.05449 30.05459 30.05468
#> 18 29.42512 29.42532 29.42549 29.42566 29.42581 29.42594 29.42607 29.42618
#> 19 17.74207 17.74197 17.74189 17.74181 17.74174 17.74167 17.74161 17.74156
#> 20 19.43458 19.43440 19.43423 19.43408 19.43394 19.43381 19.43370 19.43359
#> 21 14.84172 14.84197 14.84220 14.84241 14.84260 14.84278 14.84294 14.84308
#>      0.0552   0.0515   0.0480   0.0448   0.0417   0.0389   0.0363   0.0339
#> 1  21.33326 21.33315 21.33305 21.33296 21.33287 21.33279 21.21470 21.21433
#> 2  16.81146 16.81144 16.81143 16.81142 16.81141 16.81140 16.83517 16.83513
#> 3  13.95730 13.95724 13.95718 13.95713 13.95708 13.95703 13.78995 13.78952
#> 4  22.62611 22.62601 22.62592 22.62584 22.62577 22.62570 22.75452 22.75446
#> 5  23.36227 23.36219 23.36212 23.36206 23.36200 23.36195 23.54554 23.54554
#> 6  17.86414 17.86417 17.86419 17.86422 17.86424 17.86426 17.81918 17.81899
#> 7  15.38654 15.38655 15.38657 15.38658 15.38660 15.38661 15.37746 15.37752
#> 8  15.88003 15.88005 15.88007 15.88008 15.88010 15.88012 15.86582 15.86585
#> 9  15.99129 15.99133 15.99136 15.99138 15.99141 15.99143 15.99078 15.99087
#> 10 13.53181 13.53188 13.53195 13.53201 13.53206 13.53211 13.52219 13.52250
#> 11 12.45631 12.45629 12.45627 12.45626 12.45625 12.45623 12.39215 12.39211
#> 12 30.26601 30.26609 30.26615 30.26621 30.26627 30.26632 30.15822 30.15833
#> 13 24.50038 24.50028 24.50019 24.50010 24.50002 24.49995 24.32299 24.32260
#> 14 17.58960 17.58968 17.58975 17.58982 17.58988 17.58993 17.66442 17.66478
#> 15 16.93733 16.93732 16.93732 16.93731 16.93730 16.93730 16.96932 16.96931
#> 16 16.50688 16.50688 16.50689 16.50689 16.50690 16.50690 16.55431 16.55440
#> 17 30.05477 30.05485 30.05492 30.05499 30.05505 30.05510 29.93913 29.93931
#> 18 29.42629 29.42639 29.42648 29.42656 29.42663 29.42670 29.55956 29.56001
#> 19 17.74151 17.74146 17.74142 17.74138 17.74135 17.74132 17.90437 17.90418
#> 20 19.43349 19.43340 19.43332 19.43325 19.43318 19.43312 19.50883 19.50870
#> 21 14.84322 14.84334 14.84345 14.84355 14.84365 14.84373 14.81141 14.81156
#>      0.0316   0.0294   0.0275   0.0256   0.0239   0.0223   0.0208   0.0194
#> 1  21.36833 21.36806 21.36782 21.36760 21.36739 21.36721 21.36705 21.36690
#> 2  16.81715 16.81712 16.81708 16.81705 16.81701 16.81698 16.81695 16.81692
#> 3  13.67198 13.67198 13.67198 13.67199 13.67201 13.67203 13.67204 13.67207
#> 4  22.82949 22.82907 22.82869 22.82833 22.82801 22.82772 22.82745 22.82720
#> 5  23.41343 23.41322 23.41301 23.41282 23.41264 23.41247 23.41231 23.41216
#> 6  18.05110 18.05110 18.05109 18.05108 18.05107 18.05105 18.05102 18.05100
#> 7  15.40404 15.40408 15.40412 15.40415 15.40418 15.40420 15.40423 15.40425
#> 8  15.87484 15.87493 15.87500 15.87506 15.87512 15.87517 15.87521 15.87525
#> 9  15.98876 15.98887 15.98898 15.98907 15.98915 15.98922 15.98928 15.98933
#> 10 13.52933 13.52950 13.52965 13.52980 13.52993 13.53006 13.53017 13.53028
#> 11 12.33846 12.33844 12.33844 12.33844 12.33845 12.33846 12.33848 12.33850
#> 12 30.15348 30.15375 30.15401 30.15424 30.15446 30.15465 30.15483 30.15500
#> 13 24.17268 24.17258 24.17250 24.17243 24.17237 24.17233 24.17229 24.17227
#> 14 17.73243 17.73255 17.73265 17.73274 17.73282 17.73289 17.73295 17.73300
#> 15 17.03439 17.03434 17.03429 17.03423 17.03418 17.03413 17.03408 17.03403
#> 16 16.54625 16.54623 16.54621 16.54619 16.54617 16.54615 16.54613 16.54612
#> 17 29.96091 29.96116 29.96140 29.96162 29.96183 29.96202 29.96219 29.96236
#> 18 29.51938 29.51952 29.51965 29.51977 29.51988 29.51998 29.52007 29.52015
#> 19 17.82291 17.82273 17.82256 17.82241 17.82227 17.82214 17.82202 17.82191
#> 20 19.58752 19.58721 19.58692 19.58665 19.58641 19.58619 19.58599 19.58580
#> 21 14.68315 14.68358 14.68397 14.68433 14.68467 14.68497 14.68525 14.68550
#>      0.0181   0.0168   0.0157   0.0147   0.0137   0.0127   0.0119   0.0111
#> 1  21.36677 21.36665 21.36655 21.36645 21.36637 21.36630 21.36624 21.36618
#> 2  16.81690 16.81687 16.81685 16.81682 16.81680 16.81678 16.81676 16.81674
#> 3  13.67209 13.67211 13.67214 13.67217 13.67219 13.67222 13.67224 13.67227
#> 4  22.82698 22.82678 22.82660 22.82643 22.82628 22.82615 22.82603 22.82592
#> 5  23.41202 23.41189 23.41177 23.41166 23.41155 23.41146 23.41137 23.41129
#> 6  18.05097 18.05095 18.05092 18.05089 18.05086 18.05084 18.05081 18.05078
#> 7  15.40426 15.40428 15.40429 15.40430 15.40431 15.40432 15.40433 15.40433
#> 8  15.87528 15.87531 15.87533 15.87535 15.87536 15.87538 15.87539 15.87539
#> 9  15.98938 15.98941 15.98945 15.98947 15.98950 15.98952 15.98953 15.98954
#> 10 13.53038 13.53047 13.53056 13.53063 13.53070 13.53077 13.53083 13.53088
#> 11 12.33852 12.33855 12.33858 12.33860 12.33863 12.33866 12.33869 12.33872
#> 12 30.15515 30.15528 30.15541 30.15552 30.15562 30.15572 30.15580 30.15587
#> 13 24.17225 24.17224 24.17223 24.17223 24.17223 24.17224 24.17225 24.17226
#> 14 17.73304 17.73308 17.73311 17.73313 17.73315 17.73317 17.73318 17.73319
#> 15 17.03399 17.03394 17.03390 17.03386 17.03382 17.03378 17.03374 17.03371
#> 16 16.54610 16.54609 16.54607 16.54606 16.54605 16.54604 16.54603 16.54602
#> 17 29.96251 29.96265 29.96278 29.96290 29.96300 29.96310 29.96319 29.96328
#> 18 29.52023 29.52030 29.52036 29.52042 29.52047 29.52052 29.52056 29.52060
#> 19 17.82181 17.82172 17.82164 17.82157 17.82150 17.82144 17.82139 17.82134
#> 20 19.58564 19.58549 19.58535 19.58523 19.58512 19.58502 19.58493 19.58485
#> 21 14.68573 14.68594 14.68613 14.68630 14.68646 14.68660 14.68672 14.68683
#>      0.0103   0.0096   0.0090   0.0084   0.0078   0.0073   0.0068   0.0063
#> 1  21.36599 21.33821 21.33815 21.33808 21.33802 21.33796 21.33791 21.33785
#> 2  16.81638 16.74761 16.74746 16.74732 16.74718 16.74705 16.74693 16.74681
#> 3  13.67244 13.62246 13.62221 13.62198 13.62176 13.62155 13.62134 13.62115
#> 4  22.82574 22.83288 22.83300 22.83311 22.83322 22.83332 22.83341 22.83350
#> 5  23.41112 23.38779 23.38770 23.38761 23.38752 23.38744 23.38737 23.38729
#> 6  18.05111 18.03077 18.03091 18.03106 18.03120 18.03133 18.03146 18.03159
#> 7  15.40497 15.53435 15.53465 15.53494 15.53521 15.53547 15.53572 15.53596
#> 8  15.87601 15.93552 15.93562 15.93573 15.93583 15.93592 15.93601 15.93610
#> 9  15.99017 16.04673 16.04682 16.04691 16.04700 16.04708 16.04716 16.04723
#> 10 13.53040 13.44556 13.44535 13.44516 13.44498 13.44480 13.44463 13.44447
#> 11 12.33831 12.39116 12.39127 12.39138 12.39149 12.39158 12.39168 12.39176
#> 12 30.15552 30.13648 30.13643 30.13638 30.13633 30.13629 30.13625 30.13621
#> 13 24.17248 24.20312 24.20306 24.20299 24.20293 24.20287 24.20281 24.20275
#> 14 17.73348 17.72596 17.72598 17.72599 17.72600 17.72602 17.72603 17.72604
#> 15 17.03370 17.02111 17.02113 17.02115 17.02117 17.02119 17.02121 17.02123
#> 16 16.54544 16.47457 16.47444 16.47432 16.47420 16.47408 16.47397 16.47387
#> 17 29.96300 29.98097 29.98103 29.98108 29.98113 29.98118 29.98123 29.98127
#> 18 29.52059 29.52492 29.52495 29.52498 29.52501 29.52504 29.52506 29.52509
#> 19 17.82011 17.84027 17.84045 17.84061 17.84077 17.84092 17.84106 17.84120
#> 20 19.58541 19.59882 19.59881 19.59880 19.59879 19.59877 19.59876 19.59875
#> 21 14.68764 14.68074 14.68058 14.68042 14.68027 14.68013 14.68000 14.67987
#>      0.0059   0.0055   0.0051   0.0048
#> 1  21.33779 21.33774 21.33769 21.33764
#> 2  16.74670 16.74659 16.74649 16.74639
#> 3  13.62096 13.62078 13.62061 13.62044
#> 4  22.83358 22.83366 22.83374 22.83380
#> 5  23.38722 23.38715 23.38708 23.38701
#> 6  18.03171 18.03183 18.03194 18.03205
#> 7  15.53618 15.53640 15.53661 15.53680
#> 8  15.93618 15.93626 15.93634 15.93641
#> 9  16.04730 16.04737 16.04744 16.04750
#> 10 13.44432 13.44418 13.44405 13.44392
#> 11 12.39185 12.39193 12.39200 12.39207
#> 12 30.13617 30.13614 30.13611 30.13608
#> 13 24.20269 24.20264 24.20258 24.20253
#> 14 17.72606 17.72607 17.72609 17.72610
#> 15 17.02125 17.02127 17.02129 17.02130
#> 16 16.47377 16.47368 16.47359 16.47351
#> 17 29.98132 29.98136 29.98140 29.98145
#> 18 29.52511 29.52514 29.52516 29.52519
#> 19 17.84133 17.84145 17.84157 17.84168
#> 20 19.59873 19.59872 19.59870 19.59868
#> 21 14.67975 14.67964 14.67953 14.67943
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 18.7 14.3 24.4 22.8 17.8 16.4 17.3 15.2 10.4 14.7 33.9 21.5 15.5 15.2
#> [16] 19.2 27.3 30.4 15.8 19.7 15.0
#> 
#> $X
#>               am        carb        cyl       disp       drat       gear
#>  [1,]  1.4142136  0.56842042 -0.3244428 -0.7502198  0.9319867  0.4291975
#>  [2,] -0.7071068 -0.62526247  0.8111071  0.9092088 -0.7748494 -0.8583951
#>  [3,] -0.7071068  0.56842042  0.8111071  0.9092088 -0.6383025 -0.8583951
#>  [4,] -0.7071068 -0.62526247 -1.4599928 -0.8605718  0.4540726  0.4291975
#>  [5,] -0.7071068 -0.62526247 -1.4599928 -0.9095249  0.9775023  0.4291975
#>  [6,] -0.7071068  0.56842042 -0.3244428 -0.6871615  0.9775023  0.4291975
#>  [7,] -0.7071068 -0.02842102  0.8111071  0.2105894 -0.9569119 -0.8583951
#>  [8,] -0.7071068 -0.02842102  0.8111071  0.2105894 -0.9569119 -0.8583951
#>  [9,] -0.7071068 -0.02842102  0.8111071  0.2105894 -0.9569119 -0.8583951
#> [10,] -0.7071068  0.56842042  0.8111071  1.8384889 -1.2755213 -0.8583951
#> [11,] -0.7071068  0.56842042  0.8111071  1.5729803 -0.5927869 -0.8583951
#> [12,]  1.4142136 -1.22210391 -1.4599928 -1.4878358  1.6602367  0.4291975
#> [13,] -0.7071068 -1.22210391 -1.4599928 -1.0812758  0.4768304 -0.8583951
#> [14,] -0.7071068 -0.62526247  0.8111071  0.5607288 -1.6624041 -0.8583951
#> [15,] -0.7071068 -0.62526247  0.8111071  0.4445688 -0.7748494 -0.8583951
#> [16,] -0.7071068 -0.62526247  0.8111071  1.2410946 -0.9341541 -0.8583951
#> [17,]  1.4142136 -1.22210391 -1.4599928 -1.4222884  1.3416273  0.4291975
#> [18,]  1.4142136 -0.62526247 -1.4599928 -1.2887044  0.6361351  1.7167902
#> [19,]  1.4142136  0.56842042  0.8111071  0.8345346  1.6602367  1.7167902
#> [20,]  1.4142136  1.76210331 -0.3244428 -0.8746769  0.2947679  1.7167902
#> [21,]  1.4142136  2.95578620  0.8111071  0.4196774  0.1127054  1.7167902
#>               hp        qsec         vs         wt
#>  [1,] -0.7371632 -0.35352752 -0.7071068 -0.4906640
#>  [2,]  0.2076516 -0.35352752 -0.7071068  0.1206694
#>  [3,]  1.2251444 -0.97089269 -0.7071068  0.2613302
#>  [4,] -1.4348725  1.20558115  1.4142136 -0.1498321
#>  [5,] -0.9551973  2.72283455  1.4142136 -0.1931123
#>  [6,] -0.5482002  0.63007124  1.4142136  0.1206694
#>  [7,]  0.2803297 -0.15471500 -0.7071068  0.8023333
#>  [8,]  0.2803297 -0.05007684 -0.7071068  0.4344512
#>  [9,]  0.2803297  0.15919949 -0.7071068  0.4885515
#> [10,]  0.6437199  0.14873568 -0.7071068  2.0791005
#> [11,]  1.0071102 -0.14425119 -0.7071068  2.1818911
#> [12,] -1.3912657  1.15326207  1.4142136 -1.6159504
#> [13,] -0.9261261  1.21081306  1.4142136 -0.9342865
#> [14,] -0.1557387 -0.43200614 -0.7071068  0.2072299
#> [15,] -0.1557387 -0.20703408 -0.7071068  0.1152594
#> [16,]  0.2076516 -0.33783179 -0.7071068  0.5588819
#> [17,] -1.3767301  0.63007124  1.4142136 -1.5077498
#> [18,] -0.6935563 -0.41631042  1.4142136 -1.9643563
#> [19,]  1.5013210 -1.67196840 -0.7071068 -0.1714722
#> [20,]  0.2076516 -1.14877757 -0.7071068 -0.6042747
#> [21,]  2.5333494 -1.61964932 -0.7071068  0.2613302
#> attr(,"center")
#>  [1]   0.3333333   3.0476190   6.5714286 250.4190476   3.4904762   3.6666667
#>  [7] 160.7142857  17.6957143   0.3333333   3.3284762
#> attr(,"scale")
#>  [1]   0.4714045   1.6754869   1.7612611 120.5234125   0.4394095   0.7766432
#>  [7]  68.7965550   1.9113485   0.4714045   0.9242092
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 28
#> 
#> $lambda.min
#> [1] 0.7293543
#> 
#> $null.dev
#> [1] 31.69769
#> 
#> $Bias
#> [1] 8.281631
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 6.884283