Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 29.692974 27.464908 24.736233 21.880428 19.178131 16.911664 15.018606
#>   [8] 13.445052 12.144469 11.076720 10.207202  9.506109  8.947785  8.510166
#>  [15]  8.174297  7.921333  7.549915  7.309533  7.127250  7.109425  7.109425
#>  [22]  7.109425  7.109425  7.109425  7.080653  6.665249  6.420267  6.366523
#>  [29]  6.123506  5.749347  5.455845  5.296504  5.185545  5.274078  4.975144
#>  [36]  5.137764  4.581024  4.559681  4.540270  4.522461  4.506306  4.445231
#>  [43]  3.644823  4.568914  4.599846  4.729275  5.252032  5.427255  5.620542
#>  [50]  6.514487  7.335464  7.666802  7.508521  8.380087  8.631142  8.511850
#>  [57]  9.634546 14.625779 14.585389 14.316919 14.305176 14.305737 14.596496
#>  [64] 14.596551 14.596660 14.596821 14.597079 20.181148 20.182543 20.282975
#>  [71] 20.303872 20.305436 20.306982 20.308509 20.293998 20.179705 19.947399
#>  [78] 19.948495 19.828937 19.830072 19.831051 19.745025 19.746183 19.747328
#>  [85] 19.748461 19.749651 20.414276 20.400607 20.400988 20.401346 20.401682
#>  [92] 20.401996 20.402288 20.402560 20.402811 20.403042 20.403254 20.403446
#>  [99] 20.403594 20.401384
#> 
#> $cvse
#>   [1] 7.860836 7.813937 7.615856 7.090839 6.425106 5.828358 5.292799 4.812142
#>   [9] 4.381375 3.996563 3.654665 3.353368 3.090911 2.865900 2.677094 2.522935
#>  [17] 2.368090 2.234876 2.121942 2.111039 2.111039 2.111039 2.111039 2.111039
#>  [25] 2.105597 2.054083 2.049413 2.050785 1.891618 1.674276 1.506343 1.389733
#>  [33] 1.337720 1.336418 1.342682 1.407965 1.335779 1.326827 1.318771 1.311446
#>  [41] 1.304873 1.301142 1.189315 1.358893 1.361250 1.374601 1.431529 1.423567
#>  [49] 1.401180 1.577597 1.693358 1.642178 1.671682 2.295572 2.368606 2.387958
#>  [57] 2.583155 5.256216 5.283263 5.220108 5.222326 5.222819 5.208459 5.208975
#>  [65] 5.209488 5.209998 5.210511 7.917509 7.918313 7.907801 7.907367 7.908128
#>  [73] 7.908882 7.909627 7.911429 7.903174 7.927700 7.928441 7.934080 7.934809
#>  [81] 7.935552 7.927525 7.928231 7.928928 7.929617 7.930293 8.351835 8.352789
#>  [89] 8.352903 8.353000 8.353081 8.353144 8.353192 8.353224 8.353241 8.353244
#>  [97] 8.353232 8.353206 8.353170 8.353399
#> 
#> $fold
#>  [1]  2  8  8  4  1  6 10  5  4  7  9  1  6  3  2  3  9  7 10  5  1
#> 
#> $lambda
#>   [1] 4.916722467 4.585351828 4.276314462 3.988105179 3.719320238 3.468650502
#>   [7] 3.234875067 3.016855315 2.813529363 2.623906901 2.447064358 2.282140410
#>  [13] 2.128331784 1.984889345 1.851114447 1.726355529 1.610004944 1.501495999
#>  [19] 1.400300194 1.305924648 1.217909697 1.135826660 1.059275744 0.987884105
#>  [25] 0.921304022 0.859211215 0.801303254 0.747298097 0.696932706 0.649961774
#>  [31] 0.606156526 0.565303605 0.527204034 0.491672247 0.458535183 0.427631446
#>  [37] 0.398810518 0.371932024 0.346865050 0.323487507 0.301685531 0.281352936
#>  [43] 0.262390690 0.244706436 0.228214041 0.212833178 0.198488935 0.185111445
#>  [49] 0.172635553 0.161000494 0.150149600 0.140030019 0.130592465 0.121790970
#>  [55] 0.113582666 0.105927574 0.098788410 0.092130402 0.085921121 0.080130325
#>  [61] 0.074729810 0.069693271 0.064996177 0.060615653 0.056530361 0.052720403
#>  [67] 0.049167225 0.045853518 0.042763145 0.039881052 0.037193202 0.034686505
#>  [73] 0.032348751 0.030168553 0.028135294 0.026239069 0.024470644 0.022821404
#>  [79] 0.021283318 0.019848893 0.018511144 0.017263555 0.016100049 0.015014960
#>  [85] 0.014003002 0.013059246 0.012179097 0.011358267 0.010592757 0.009878841
#>  [91] 0.009213040 0.008592112 0.008013033 0.007472981 0.006969327 0.006499618
#>  [97] 0.006061565 0.005653036 0.005272040 0.004916722
#> 
#> $fit
#> $beta
#>               4.9167       4.5854      4.2763      3.9881      3.7193
#> (Intercept) 18.67619 19.676879310 20.61012506 21.48047318 22.29216274
#> am           0.00000  0.000000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000  0.000000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000  0.000000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000 -0.004025528 -0.00777975 -0.01128095 -0.01454618
#> drat         0.00000  0.000000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000  0.000000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000  0.000000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000  0.000000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000  0.000000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000  0.000000000  0.00000000  0.00000000  0.00000000
#>                  3.4687      3.2349     3.0169      2.8135      2.6239
#> (Intercept) 23.04914715 23.75511334 24.4134998 25.02751316 25.60014410
#> am           0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> disp        -0.01759134 -0.02043127 -0.0230798 -0.02554983 -0.02785339
#> drat         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#>                  2.4471     2.2821      2.1283      1.9849      1.8511
#> (Intercept) 26.13418163 26.6322268 27.09670544 27.52987974 27.93385955
#> am           0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> disp        -0.03000169 -0.0320052 -0.03387369 -0.03561624 -0.03724136
#> drat         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.0000000  0.00000000  0.00000000  0.00000000
#>                  1.7264      1.6100      1.5015      1.4003      1.3059
#> (Intercept) 28.31061247 28.57469223 28.57469223 28.57469223 28.57469223
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.03875694 -0.03981927 -0.03981927 -0.03981927 -0.03981927
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  1.2179      1.1358      1.0593      0.9879      0.9213
#> (Intercept) 28.57469223 28.57469223 28.57469223 28.57469223 28.57469223
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.03981927 -0.03981927 -0.03981927 -0.03981927 -0.03981927
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.8592      0.8013      0.7473      0.6969      0.6500
#> (Intercept) 28.57469223 28.57469223 28.57469223 28.57469223 28.57469223
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.03981927 -0.03981927 -0.03981927 -0.03981927 -0.03981927
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.6062      0.5653      0.5272      0.4917      0.4585
#> (Intercept) 28.57469223 28.57469223 28.57469223 28.64622956 19.67769454
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000 -0.04546490 -1.05411155
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.03981927 -0.03981927 -0.03981927 -0.03953224 -0.02475325
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  2.25011628
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.4276      0.3988      0.3719      0.3469      0.3235
#> (Intercept) 19.67634016 19.67537741 19.67469304 19.67420655 19.67386074
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb        -1.05429504 -1.05442547 -1.05451819 -1.05458410 -1.05463095
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.02475039 -0.02474836 -0.02474691 -0.02474588 -0.02474515
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         2.25044040  2.25067080  2.25083458  2.25095100  2.25103376
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.3017      0.2814    0.2624      0.2447      0.2282
#> (Intercept) 19.67361491 19.67344017 19.673316 19.67322766 19.67316490
#> am           0.00000000  0.00000000  0.000000  0.00000000  0.00000000
#> carb        -1.05466425 -1.05468792 -1.054705 -1.05471671 -1.05472522
#> cyl          0.00000000  0.00000000  0.000000  0.00000000  0.00000000
#> disp        -0.02474463 -0.02474426 -0.024744 -0.02474381 -0.02474368
#> drat         0.00000000  0.00000000  0.000000  0.00000000  0.00000000
#> gear         2.25109259  2.25113441  2.251164  2.25118526  2.25120029
#> hp           0.00000000  0.00000000  0.000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.000000  0.00000000  0.00000000
#>                  0.2128      0.1985      0.1851      0.1726      0.1610
#> (Intercept) 19.67312028 19.67308856 19.67306602 19.67304999 18.32961533
#> am           0.00000000  0.00000000  0.00000000  0.00000000 -1.24350423
#> carb        -1.05473126 -1.05473556 -1.05473861 -1.05474078 -1.03760289
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.02474358 -0.02474352 -0.02474347 -0.02474343 -0.02573606
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         2.25121096  2.25121855  2.25122395  2.25122778  2.80128298
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                 0.1501     0.1400      0.1306      0.1218      0.1136
#> (Intercept) 18.3284884 18.3277061 18.32716982 37.25996138 37.26419627
#> am          -1.2434758 -1.2434264 -1.24336973 -2.55646811 -2.55725738
#> carb        -1.0377557 -1.0378674 -1.03794846 -0.84685145 -0.84674378
#> cyl          0.0000000  0.0000000  0.00000000 -1.09975545 -1.09993343
#> disp        -0.0257336 -0.0257318 -0.02573049 -0.02145251 -0.02145405
#> drat         0.0000000  0.0000000  0.00000000  0.00000000  0.00000000
#> gear         2.8015445  2.8017213  2.80183874  2.11664658  2.11656494
#> hp           0.0000000  0.0000000  0.00000000  0.00000000  0.00000000
#> qsec         0.0000000  0.0000000  0.00000000 -0.59532919 -0.59546672
#> vs           0.0000000  0.0000000  0.00000000  0.00000000  0.00000000
#> wt           0.0000000  0.0000000  0.00000000  0.00000000  0.00000000
#>                  0.1059      0.0988      0.0921      0.0859      0.0801
#> (Intercept) 37.26762378 37.27038290 37.27259065 37.27434524 40.00158148
#> am          -2.55791993 -2.55847458 -2.55893759 -2.55932296 -2.44850214
#> carb        -0.84665748 -0.84658878 -0.84653453 -0.84649207 -0.85750984
#> cyl         -1.10007410 -1.10018431 -1.10026978 -1.10033525 -1.20634210
#> disp        -0.02145536 -0.02145647 -0.02145742 -0.02145823 -0.02091523
#> drat         0.00000000  0.00000000  0.00000000  0.00000000 -0.57017670
#> gear         2.11650792  2.11647016  2.11644730  2.11643582  2.22793928
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec        -0.59557947 -0.59567153 -0.59574637 -0.59580691 -0.63059602
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.0747     0.0697      0.0650      0.0606     0.0565
#> (Intercept) 40.00568719 40.0091022 40.01192162 40.01422978 40.0161015
#> am          -2.44907914 -2.4495923 -2.45004654 -2.45044667 -2.4507975
#> carb        -0.85736390 -0.8572395 -0.85713398 -0.85704507 -0.8569706
#> cyl         -1.20657134 -1.2067606 -1.20691540 -1.20704091 -1.2071415
#> disp        -0.02091609 -0.0209169 -0.02091766 -0.02091836 -0.0209190
#> drat        -0.57000399 -0.5698430 -0.56969436 -0.56955804 -0.5694339
#> gear         2.22766307  2.2274327  2.22724208  2.22708560  2.2269584
#> hp           0.00000000  0.0000000  0.00000000  0.00000000  0.0000000
#> qsec        -0.63072200 -0.6308291 -0.63091952 -0.63099552 -0.6310590
#> vs           0.00000000  0.0000000  0.00000000  0.00000000  0.0000000
#> wt           0.00000000  0.0000000  0.00000000  0.00000000  0.0000000
#>                  0.0527      0.0492     0.0459      0.0428      0.0399
#> (Intercept) 40.01760240 40.01879027 40.0197154 40.02042133 40.02094590
#> am          -2.45110373 -2.45136972 -2.4515997 -2.45179757 -2.45196697
#> carb        -0.85690871 -0.85685762 -0.8568158 -0.85678196 -0.85675483
#> cyl         -1.20722104 -1.20728292 -1.2073301 -1.20736505 -1.20739001
#> disp        -0.02091959 -0.02092012 -0.0209206 -0.02092102 -0.02092141
#> drat        -0.56932163 -0.56922064 -0.5691303 -0.56904997 -0.56897885
#> gear         2.22685605  2.22677482  2.2267113  2.22666269  2.22662635
#> hp           0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> qsec        -0.63111152 -0.63115474 -0.6311900 -0.63121836 -0.63124100
#> vs           0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#>                  0.0372      0.0347      0.0323       0.0302       0.0281
#> (Intercept) 39.67155539 39.67415521 39.67661374 38.582791636 38.585272897
#> am          -2.39001611 -2.38990283 -2.38979610 -2.365948801 -2.365765709
#> carb        -0.71329006 -0.71313269 -0.71298394 -0.591244190 -0.591117607
#> cyl         -1.26433844 -1.26457225 -1.26479327 -1.162557344 -1.162807135
#> disp        -0.01597801 -0.01597401 -0.01597022 -0.013548152 -0.013544099
#> drat        -0.59654716 -0.59665513 -0.59675706 -0.478077598 -0.478242093
#> gear         2.00012729  1.99977546  1.99944293  2.012587551  2.012228670
#> hp           0.00000000  0.00000000  0.00000000 -0.005658742 -0.005657189
#> qsec        -0.50081178 -0.50078732 -0.50076429 -0.469509997 -0.469466657
#> vs           0.00000000  0.00000000  0.00000000  0.000000000  0.000000000
#> wt          -0.68814568 -0.68855088 -0.68893363 -0.877783427 -0.878213390
#>                   0.0262       0.0245       0.0228       0.0213       0.0198
#> (Intercept) 38.587692069 38.590048677 38.592342428 38.594573192 38.596740993
#> am          -2.365594123 -2.365433507 -2.365283341 -2.365143126 -2.365012378
#> carb        -0.590998049 -0.590885222 -0.590778840 -0.590678625 -0.590584308
#> cyl         -1.163048357 -1.163281155 -1.163505678 -1.163722086 -1.163930540
#> disp        -0.013540286 -0.013536703 -0.013533338 -0.013530183 -0.013527226
#> drat        -0.478400211 -0.478552107 -0.478697939 -0.478837863 -0.478972038
#> gear         2.011885104  2.011556400  2.011242104  2.010941770  2.010654953
#> hp          -0.005655645 -0.005654113 -0.005652597 -0.005651097 -0.005649617
#> qsec        -0.469427359 -0.469391865 -0.469359951 -0.469331400 -0.469306008
#> vs           0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> wt          -0.878617840 -0.878997915 -0.879354715 -0.879689304 -0.880002711
#>                   0.0185       0.0173      0.0161       0.0150       0.0140
#> (Intercept) 38.003505837 38.002338666 38.00123654 38.000197589 37.999219989
#> am          -2.453135836 -2.452978114 -2.45282072 -2.452663764 -2.452507349
#> carb        -0.549101339 -0.548968216 -0.54883875 -0.548712891 -0.548590585
#> cyl         -1.040384998 -1.040340270 -1.04030185 -1.040269499 -1.040242991
#> disp        -0.013637146 -0.013632905 -0.01362874 -0.013624658 -0.013620653
#> drat        -0.329459908 -0.329422603 -0.32939150 -0.329366370 -0.329346977
#> gear         2.051229725  2.051122730  2.05101475  2.050905907  2.050796324
#> hp          -0.008137559 -0.008140655 -0.00814359 -0.008146369 -0.008148996
#> qsec        -0.518549614 -0.518458757 -0.51836870 -0.518279484 -0.518191140
#> vs           0.401927912  0.402010183  0.40208026  0.402138590  0.402185627
#> wt          -0.821441364 -0.821884048 -0.82231972 -0.822748332 -0.823169847
#>                   0.0131       0.0122       0.0114       0.0106       0.0099
#> (Intercept) 37.998301933 37.997441647 37.996637386 37.995887433 37.995190102
#> am          -2.452351574 -2.452196534 -2.452042318 -2.451889013 -2.451736699
#> carb        -0.548471778 -0.548356416 -0.548244447 -0.548135814 -0.548030463
#> cyl         -1.040222104 -1.040206617 -1.040196317 -1.040190995 -1.040190447
#> disp        -0.013616726 -0.013612878 -0.013609108 -0.013605415 -0.013601800
#> drat        -0.329333099 -0.329324515 -0.329321011 -0.329322378 -0.329328410
#> gear         2.050686117  2.050575397  2.050464273  2.050352848  2.050241220
#> hp          -0.008151476 -0.008153812 -0.008156009 -0.008158072 -0.008160003
#> qsec        -0.518103700 -0.518017196 -0.517931656 -0.517847105 -0.517763569
#> vs           0.402221803  0.402247546  0.402263271  0.402269384  0.402266283
#> wt          -0.823584233 -0.823991466 -0.824391528 -0.824784407 -0.825170097
#>                   0.0092       0.0086       0.0080       0.0075      0.0070
#> (Intercept) 37.994543736 37.993946707 37.993397417 37.992894298 37.99243581
#> am          -2.451585454 -2.451435351 -2.451286459 -2.451138843 -2.45099257
#> carb        -0.547928339 -0.547829386 -0.547733549 -0.547640772 -0.54755100
#> cyl         -1.040194475 -1.040202882 -1.040215481 -1.040232086 -1.04025252
#> disp        -0.013598262 -0.013594801 -0.013591415 -0.013588105 -0.01358487
#> drat        -0.329338907 -0.329353676 -0.329372526 -0.329395271 -0.32942173
#> gear         2.050129486  2.050017737  2.049906061  2.049794541  2.04968326
#> hp          -0.008161807 -0.008163488 -0.008165049 -0.008166495 -0.00816783
#> qsec        -0.517681070 -0.517599628 -0.517519263 -0.517439994 -0.51736183
#> vs           0.402254355  0.402233979  0.402205523  0.402169348  0.40212580
#> wt          -0.825548599 -0.825919918 -0.826284067 -0.826641060 -0.82699092
#>                   0.0065       0.0061       0.0057      0.0053       0.0049
#> (Intercept) 37.992020450 37.991646732 37.991313208 37.99101846 37.990761088
#> am          -2.450847683 -2.450704251 -2.450562321 -2.45042194 -2.450283151
#> carb        -0.547464176 -0.547380246 -0.547299153 -0.54722084 -0.547145257
#> cyl         -1.040276596 -1.040304154 -1.040335023 -1.04036904 -1.040406051
#> disp        -0.013581709 -0.013578621 -0.013575607 -0.01357266 -0.013569794
#> drat        -0.329451730 -0.329485097 -0.329521665 -0.32956127 -0.329603760
#> gear         2.049572288  2.049461704  2.049351578  2.04924197  2.049132958
#> hp          -0.008169056 -0.008170177 -0.008171197 -0.00817212 -0.008172948
#> qsec        -0.517284800 -0.517208903 -0.517134156 -0.51706057 -0.516988149
#> vs           0.402075234  0.402017970  0.401954338  0.40188465  0.401809220
#> wt          -0.827333671 -0.827669345 -0.827997975 -0.82831960 -0.828634266
#> 
#> $iter
#>   [1]   0   5   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   7  41   1
#>  [37]   1   1   1   1   1   1   1   1   1   1   1   1   1  49   1   1   1  53
#>  [55]   1   1   1   1   1  77   1   1   1   1   1   1   1   1   1   1  94   1
#>  [73]   1 110   1   1   1   1   1   1 128   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.916722467 4.585351828 4.276314462 3.988105179 3.719320238 3.468650502
#>   [7] 3.234875067 3.016855315 2.813529363 2.623906901 2.447064358 2.282140410
#>  [13] 2.128331784 1.984889345 1.851114447 1.726355529 1.610004944 1.501495999
#>  [19] 1.400300194 1.305924648 1.217909697 1.135826660 1.059275744 0.987884105
#>  [25] 0.921304022 0.859211215 0.801303254 0.747298097 0.696932706 0.649961774
#>  [31] 0.606156526 0.565303605 0.527204034 0.491672247 0.458535183 0.427631446
#>  [37] 0.398810518 0.371932024 0.346865050 0.323487507 0.301685531 0.281352936
#>  [43] 0.262390690 0.244706436 0.228214041 0.212833178 0.198488935 0.185111445
#>  [49] 0.172635553 0.161000494 0.150149600 0.140030019 0.130592465 0.121790970
#>  [55] 0.113582666 0.105927574 0.098788410 0.092130402 0.085921121 0.080130325
#>  [61] 0.074729810 0.069693271 0.064996177 0.060615653 0.056530361 0.052720403
#>  [67] 0.049167225 0.045853518 0.042763145 0.039881052 0.037193202 0.034686505
#>  [73] 0.032348751 0.030168553 0.028135294 0.026239069 0.024470644 0.022821404
#>  [79] 0.021283318 0.019848893 0.018511144 0.017263555 0.016100049 0.015014960
#>  [85] 0.014003002 0.013059246 0.012179097 0.011358267 0.010592757 0.009878841
#>  [91] 0.009213040 0.008592112 0.008013033 0.007472981 0.006969327 0.006499618
#>  [97] 0.006061565 0.005653036 0.005272040 0.004916722
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 34
#> 
#> $loss
#>   [1] 598.31810 500.86323 419.32775 351.42072 295.16428 248.85183 211.01181
#>   [8] 180.37621 155.85329 136.50382 121.52054 110.21036 101.97882  96.31673
#>  [15]  92.78849  91.02207  90.66074  90.66074  90.66074  90.66074  90.66074
#>  [22]  90.66074  90.66074  90.66074  90.66074  90.66074  90.66074  90.66074
#>  [29]  90.66074  90.66074  90.66074  90.66074  90.66074  89.07867  55.41988
#>  [36]  55.41987  55.41987  55.41987  55.41987  55.41987  55.41987  55.41987
#>  [43]  55.41987  55.41987  55.41987  55.41987  55.41987  55.41987  55.41987
#>  [50]  53.32488  53.32487  53.32487  53.32487  45.23775  45.23774  45.23773
#>  [57]  45.23773  45.23773  45.23773  44.76305  44.76304  44.76303  44.76303
#>  [64]  44.76303  44.76302  44.76302  44.76302  44.76302  44.76302  44.76302
#>  [71]  44.19613  44.19611  44.19610  43.92573  43.92571  43.92570  43.92569
#>  [78]  43.92567  43.92566  43.92565  43.85402  43.85401  43.85400  43.85398
#>  [85]  43.85397  43.85396  43.85395  43.85394  43.85393  43.85392  43.85391
#>  [92]  43.85390  43.85390  43.85389  43.85388  43.85387  43.85387  43.85386
#>  [99]  43.85385  43.85385
#> 
#> $linear.predictors
#>      4.9167   4.5854   4.2763   3.9881   3.7193   3.4687   3.2349   3.0169
#> 1  18.67619 19.03279 19.36537 19.67552 19.96477 20.23453 20.48611 20.72073
#> 2  18.67619 19.24212 19.76991 20.26213 20.72118 21.14928 21.54854 21.92088
#> 3  18.67619 18.77114 18.85968 18.94226 19.01927 19.09109 19.15808 19.22054
#> 4  18.67619 18.22769 17.80942 17.41933 17.05554 16.71626 16.39985 16.10477
#> 5  18.67619 19.08633 19.46884 19.82556 20.15824 20.46850 20.75785 21.02769
#> 6  18.67619 19.11008 19.51474 19.89212 20.24406 20.57229 20.87839 21.16386
#> 7  18.67619 18.56664 18.46447 18.36919 18.28033 18.19745 18.12017 18.04809
#> 8  18.67619 18.56664 18.46447 18.36919 18.28033 18.19745 18.12017 18.04809
#> 9  18.67619 18.56664 18.46447 18.36919 18.28033 18.19745 18.12017 18.04809
#> 10 18.67619 17.77683 16.93808 16.15587 15.42637 14.74603 14.11155 13.51983
#> 11 18.67619 17.82514 17.03144 16.29124 15.60092 14.95713 14.35673 13.79679
#> 12 18.67619 17.90565 17.18704 16.51686 15.89184 15.30896 14.76535 14.25839
#> 13 18.67619 18.39676 18.13616 17.89313 17.66648 17.45510 17.25797 17.07412
#> 14 18.67619 18.26794 17.88721 17.53214 17.20100 16.89218 16.60417 16.33557
#> 15 18.67619 19.35886 19.99552 20.58928 21.14301 21.65943 22.14104 22.59020
#> 16 18.67619 19.19261 19.67422 20.12338 20.54226 20.93291 21.29723 21.63700
#> 17 18.67619 19.29405 19.87027 20.40765 20.90882 21.37621 21.81210 22.21861
#> 18 18.67619 18.26392 17.87943 17.52086 17.18645 16.87459 16.58374 16.31249
#> 19 18.67619 19.09318 19.48206 19.84474 20.18297 20.49840 20.79258 21.06693
#> 20 18.67619 18.46520 18.26842 18.08491 17.91376 17.75415 17.60530 17.46648
#> 21 18.67619 19.18979 19.66878 20.11548 20.53208 20.92059 21.28293 21.62084
#>      2.8135   2.6239   2.4471   2.2821   2.1283   1.9849   1.8511   1.7264
#> 1  20.93954 21.14360 21.33391 21.51139 21.67692 21.83128 21.97524 22.10950
#> 2  22.26813 22.59198 22.89400 23.17566 23.43835 23.68333 23.91179 24.12486
#> 3  19.27880 19.33313 19.38380 19.43106 19.47513 19.51623 19.55455 19.59030
#> 4  15.82957 15.57293 15.33357 15.11035 14.90218 14.70803 14.52697 14.35811
#> 5  21.27935 21.51405 21.73293 21.93706 22.12744 22.30498 22.47055 22.62497
#> 6  21.43010 21.67839 21.90994 22.12589 22.32729 22.51511 22.69028 22.85364
#> 7  17.98087 17.91818 17.85972 17.80519 17.75434 17.70692 17.66269 17.62145
#> 8  17.98087 17.91818 17.85972 17.80519 17.75434 17.70692 17.66269 17.62145
#> 9  17.98087 17.91818 17.85972 17.80519 17.75434 17.70692 17.66269 17.62145
#> 10 12.96799 12.45335 11.97338 11.52577 11.10832 10.71901 10.35594 10.01734
#> 11 13.27459 12.78759 12.33340 11.90983 11.51481 11.14641 10.80284 10.48242
#> 12 13.78559 13.34465 12.93344 12.54994 12.19228 11.85873 11.54766 11.25756
#> 13 16.90267 16.74277 16.59364 16.45457 16.32487 16.20391 16.09111 15.98591
#> 14 16.08507 15.85146 15.63359 15.43041 15.24091 15.06419 14.89938 14.74568
#> 15 23.00908 23.39973 23.76405 24.10382 24.42068 24.71620 24.99179 25.24881
#> 16 21.95387 22.24938 22.52498 22.78200 23.02170 23.24525 23.45372 23.64815
#> 17 22.59772 22.95129 23.28102 23.58853 23.87532 24.14278 24.39221 24.62483
#> 18 16.05952 15.82361 15.60359 15.39840 15.20704 15.02858 14.86214 14.70693
#> 19 21.32279 21.56140 21.78394 21.99147 22.18502 22.36552 22.53386 22.69086
#> 20 17.33701 17.21628 17.10367 16.99866 16.90073 16.80939 16.72421 16.64477
#> 21 21.93598 22.22988 22.50398 22.75960 22.99799 23.22031 23.42766 23.62102
#>       1.6100    1.5015    1.4003    1.3059    1.2179    1.1358    1.0593
#> 1  22.203609 22.203609 22.203609 22.203609 22.203609 22.203609 22.203609
#> 2  24.274211 24.274211 24.274211 24.274211 24.274211 24.274211 24.274211
#> 3  19.615356 19.615356 19.615356 19.615356 19.615356 19.615356 19.615356
#> 4  14.239755 14.239755 14.239755 14.239755 14.239755 14.239755 14.239755
#> 5  22.733205 22.733205 22.733205 22.733205 22.733205 22.733205 22.733205
#> 6  22.968139 22.968139 22.968139 22.968139 22.968139 22.968139 22.968139
#> 7  17.592537 17.592537 17.592537 17.592537 17.592537 17.592537 17.592537
#> 8  17.592537 17.592537 17.592537 17.592537 17.592537 17.592537 17.592537
#> 9  17.592537 17.592537 17.592537 17.592537 17.592537 17.592537 17.592537
#> 10  9.779997  9.779997  9.779997  9.779997  9.779997  9.779997  9.779997
#> 11 10.257828 10.257828 10.257828 10.257828 10.257828 10.257828 10.257828
#> 12 11.054213 11.054213 11.054213 11.054213 11.054213 11.054213 11.054213
#> 13 15.912164 15.912164 15.912164 15.912164 15.912164 15.912164 15.912164
#> 14 14.637948 14.637948 14.637948 14.637948 14.637948 14.637948 14.637948
#> 15 25.428970 25.428970 25.428970 25.428970 25.428970 25.428970 25.428970
#> 16 23.784434 23.784434 23.784434 23.784434 23.784434 23.784434 23.784434
#> 17 24.787880 24.787880 24.787880 24.787880 24.787880 24.787880 24.787880
#> 18 14.598128 14.598128 14.598128 14.598128 14.598128 14.598128 14.598128
#> 19 22.800898 22.800898 22.800898 22.800898 22.800898 22.800898 22.800898
#> 20 16.589092 16.589092 16.589092 16.589092 16.589092 16.589092 16.589092
#> 21 23.756561 23.756561 23.756561 23.756561 23.756561 23.756561 23.756561
#>       0.9879    0.9213    0.8592    0.8013    0.7473    0.6969    0.6500
#> 1  22.203609 22.203609 22.203609 22.203609 22.203609 22.203609 22.203609
#> 2  24.274211 24.274211 24.274211 24.274211 24.274211 24.274211 24.274211
#> 3  19.615356 19.615356 19.615356 19.615356 19.615356 19.615356 19.615356
#> 4  14.239755 14.239755 14.239755 14.239755 14.239755 14.239755 14.239755
#> 5  22.733205 22.733205 22.733205 22.733205 22.733205 22.733205 22.733205
#> 6  22.968139 22.968139 22.968139 22.968139 22.968139 22.968139 22.968139
#> 7  17.592537 17.592537 17.592537 17.592537 17.592537 17.592537 17.592537
#> 8  17.592537 17.592537 17.592537 17.592537 17.592537 17.592537 17.592537
#> 9  17.592537 17.592537 17.592537 17.592537 17.592537 17.592537 17.592537
#> 10  9.779997  9.779997  9.779997  9.779997  9.779997  9.779997  9.779997
#> 11 10.257828 10.257828 10.257828 10.257828 10.257828 10.257828 10.257828
#> 12 11.054213 11.054213 11.054213 11.054213 11.054213 11.054213 11.054213
#> 13 15.912164 15.912164 15.912164 15.912164 15.912164 15.912164 15.912164
#> 14 14.637948 14.637948 14.637948 14.637948 14.637948 14.637948 14.637948
#> 15 25.428970 25.428970 25.428970 25.428970 25.428970 25.428970 25.428970
#> 16 23.784434 23.784434 23.784434 23.784434 23.784434 23.784434 23.784434
#> 17 24.787880 24.787880 24.787880 24.787880 24.787880 24.787880 24.787880
#> 18 14.598128 14.598128 14.598128 14.598128 14.598128 14.598128 14.598128
#> 19 22.800898 22.800898 22.800898 22.800898 22.800898 22.800898 22.800898
#> 20 16.589092 16.589092 16.589092 16.589092 16.589092 16.589092 16.589092
#> 21 23.756561 23.756561 23.756561 23.756561 23.756561 23.756561 23.756561
#>       0.6062    0.5653    0.5272    0.4917   0.4585   0.4276   0.3988   0.3719
#> 1  22.203609 22.203609 22.203609 22.139212 20.50119 20.50086 20.50062 20.50045
#> 2  24.274211 24.274211 24.274211 24.331283 24.95070 24.95076 24.95081 24.95085
#> 3  19.615356 19.615356 19.615356 19.706011 19.80445 19.80453 19.80458 19.80462
#> 4  14.239755 14.239755 14.239755 14.232765 13.30043 13.30034 13.30028 13.30024
#> 5  22.733205 22.733205 22.733205 22.755921 22.93863 22.93863 22.93863 22.93862
#> 6  22.968139 22.968139 22.968139 22.989161 23.08468 23.08466 23.08464 23.08463
#> 7  17.592537 17.592537 17.592537 17.606844 16.43876 16.43862 16.43852 16.43844
#> 8  17.592537 17.592537 17.592537 17.606844 16.43876 16.43862 16.43852 16.43844
#> 9  17.592537 17.592537 17.592537 17.606844 16.43876 16.43862 16.43852 16.43844
#> 10  9.779997  9.779997  9.779997  9.805154 10.52806 10.52830 10.52846 10.52858
#> 11 10.257828 10.257828 10.257828 10.279541 10.82510 10.82530 10.82544 10.82555
#> 12 11.054213 11.054213 11.054213 11.070186 11.32017 11.32031 11.32041 11.32048
#> 13 15.912164 15.912164 15.912164 15.984048 16.44829 16.44845 16.44856 16.44864
#> 14 14.637948 14.637948 14.637948 14.628087 13.54796 13.54784 13.54776 13.54771
#> 15 25.428970 25.428970 25.428970 25.477718 25.66854 25.66853 25.66852 25.66851
#> 16 23.784434 23.784434 23.784434 23.799572 25.84224 25.84248 25.84265 25.84278
#> 17 24.787880 24.787880 24.787880 24.795784 26.46602 26.46619 26.46631 26.46640
#> 18 14.598128 14.598128 14.598128 14.588555 18.02344 18.02398 18.02436 18.02463
#> 19 22.800898 22.800898 22.800898 22.641266 21.01438 21.01397 21.01367 21.01346
#> 20 16.589092 16.589092 16.589092 16.383307 15.04465 15.04431 15.04407 15.04390
#> 21 23.756561 23.756561 23.756561 23.771899 23.57479 23.57471 23.57466 23.57462
#>      0.3469   0.3235   0.3017   0.2814   0.2624   0.2447   0.2282   0.2128
#> 1  20.50033 20.50025 20.50019 20.50014 20.50011 20.50009 20.50008 20.50007
#> 2  24.95087 24.95089 24.95090 24.95091 24.95092 24.95092 24.95092 24.95093
#> 3  19.80465 19.80467 19.80469 19.80470 19.80470 19.80471 19.80471 19.80472
#> 4  13.30021 13.30018 13.30017 13.30016 13.30015 13.30014 13.30014 13.30014
#> 5  22.93862 22.93862 22.93862 22.93862 22.93862 22.93862 22.93862 22.93862
#> 6  23.08462 23.08462 23.08461 23.08461 23.08461 23.08461 23.08461 23.08461
#> 7  16.43839 16.43836 16.43833 16.43831 16.43830 16.43829 16.43828 16.43828
#> 8  16.43839 16.43836 16.43833 16.43831 16.43830 16.43829 16.43828 16.43828
#> 9  16.43839 16.43836 16.43833 16.43831 16.43830 16.43829 16.43828 16.43828
#> 10 10.52867 10.52873 10.52877 10.52880 10.52882 10.52884 10.52885 10.52886
#> 11 10.82562 10.82567 10.82571 10.82573 10.82575 10.82576 10.82577 10.82578
#> 12 11.32054 11.32057 11.32060 11.32062 11.32063 11.32064 11.32065 11.32065
#> 13 16.44870 16.44874 16.44877 16.44879 16.44881 16.44882 16.44883 16.44883
#> 14 13.54767 13.54764 13.54762 13.54760 13.54759 13.54758 13.54758 13.54757
#> 15 25.66850 25.66850 25.66850 25.66849 25.66849 25.66849 25.66849 25.66849
#> 16 25.84286 25.84293 25.84297 25.84300 25.84302 25.84304 25.84305 25.84306
#> 17 26.46646 26.46650 26.46654 26.46656 26.46657 26.46658 26.46659 26.46660
#> 18 18.02482 18.02496 18.02506 18.02513 18.02517 18.02521 18.02523 18.02525
#> 19 21.01330 21.01320 21.01312 21.01307 21.01303 21.01300 21.01298 21.01297
#> 20 15.04378 15.04369 15.04363 15.04359 15.04356 15.04353 15.04352 15.04351
#> 21 23.57459 23.57457 23.57456 23.57455 23.57454 23.57453 23.57453 23.57453
#>      0.1985   0.1851   0.1726   0.1610   0.1501   0.1400   0.1306   0.1218
#> 1  20.50006 20.50005 20.50005 20.02306 20.02279 20.02261 20.02248 19.95262
#> 2  24.95093 24.95093 24.95093 24.47415 24.47421 24.47426 24.47431 24.52826
#> 3  19.80472 19.80472 19.80472 19.90525 19.90531 19.90535 19.90538 19.30014
#> 4  13.30014 13.30013 13.30013 13.31807 13.31800 13.31795 13.31791 14.27153
#> 5  22.93862 22.93862 22.93862 23.68406 23.68404 23.68400 23.68396 24.58016
#> 6  23.08460 23.08460 23.08460 23.83590 23.83586 23.83582 23.83577 22.98027
#> 7  16.43828 16.43827 16.43827 16.52265 16.52253 16.52244 16.52237 15.99597
#> 8  16.43828 16.43827 16.43827 16.52265 16.52253 16.52244 16.52237 15.87691
#> 9  16.43828 16.43827 16.43827 16.52265 16.52253 16.52244 16.52237 15.63877
#> 10 10.52886 10.52887 10.52887 10.43563 10.43584 10.43599 10.43610 10.59485
#> 11 10.82578 10.82579 10.82579 10.74447 10.74464 10.74477 10.74486 10.94753
#> 12 11.32065 11.32066 11.32066 11.25919 11.25931 11.25941 11.25947 11.61471
#> 13 16.44883 16.44884 16.44884 16.47419 16.47433 16.47442 16.47449 16.25305
#> 14 13.54757 13.54757 13.54757 13.57543 13.57534 13.57527 13.57522 14.74205
#> 15 25.66849 25.66849 25.66849 25.22049 25.22048 25.22049 25.22050 24.97774
#> 16 25.84307 25.84307 25.84307 25.92127 25.92147 25.92162 25.92172 26.67127
#> 17 26.46660 26.46660 26.46661 26.56982 26.56996 26.57006 26.57013 27.09280
#> 18 18.02526 18.02527 18.02528 17.90876 17.90922 17.90955 17.90980 16.93917
#> 19 21.01296 21.01295 21.01295 21.13518 21.13483 21.13457 21.13438 21.26887
#> 20 15.04350 15.04349 15.04349 15.04515 15.04488 15.04467 15.04453 14.56486
#> 21 23.57453 23.57452 23.57452 23.10197 23.10191 23.10188 23.10187 23.40848
#>      0.1136   0.1059   0.0988   0.0921   0.0859   0.0801   0.0747   0.0697
#> 1  19.95259 19.95257 19.95254 19.95251 19.95249 19.84701 19.84710 19.84718
#> 2  24.52805 24.52787 24.52772 24.52759 24.52748 24.59254 24.59242 24.59231
#> 3  19.30005 19.29997 19.29991 19.29987 19.29983 19.55957 19.55935 19.55916
#> 4  14.27180 14.27202 14.27220 14.27234 14.27246 14.25623 14.25651 14.25674
#> 5  24.58059 24.58095 24.58125 24.58150 24.58170 24.58881 24.58918 24.58950
#> 6  22.98032 22.98036 22.98040 22.98043 22.98047 22.75234 22.75239 22.75244
#> 7  15.99605 15.99611 15.99617 15.99623 15.99627 15.97090 15.97088 15.97087
#> 8  15.87695 15.87700 15.87704 15.87708 15.87711 15.84478 15.84474 15.84471
#> 9  15.63877 15.63876 15.63877 15.63878 15.63879 15.59254 15.59245 15.59237
#> 10 10.59465 10.59448 10.59433 10.59421 10.59411 10.72390 10.72376 10.72363
#> 11 10.94737 10.94723 10.94712 10.94702 10.94694 11.03587 11.03577 11.03568
#> 12 11.61464 11.61457 11.61452 11.61447 11.61443 11.57527 11.57528 11.57529
#> 13 16.25303 16.25301 16.25300 16.25300 16.25301 16.45676 16.45657 16.45641
#> 14 14.74239 14.74267 14.74290 14.74309 14.74324 14.44005 14.44048 14.44085
#> 15 24.97753 24.97736 24.97721 24.97709 24.97698 24.88507 24.88498 24.88490
#> 16 26.67133 26.67138 26.67142 26.67145 26.67148 26.57945 26.57953 26.57960
#> 17 27.09287 27.09293 27.09298 27.09303 27.09306 27.35671 27.35667 27.35664
#> 18 16.93868 16.93829 16.93797 16.93772 16.93751 16.72097 16.72046 16.72004
#> 19 21.26913 21.26935 21.26952 21.26967 21.26979 21.43868 21.43887 21.43904
#> 20 14.56486 14.56487 14.56487 14.56487 14.56487 14.66135 14.66134 14.66134
#> 21 23.40836 23.40825 23.40815 23.40806 23.40799 23.32119 23.32125 23.32130
#>      0.0650   0.0606   0.0565   0.0527   0.0492   0.0459   0.0428   0.0399
#> 1  19.84725 19.84730 19.84734 19.84738 19.84741 19.84743 19.84744 19.84746
#> 2  24.59220 24.59211 24.59202 24.59194 24.59187 24.59180 24.59175 24.59169
#> 3  19.55900 19.55886 19.55873 19.55863 19.55854 19.55846 19.55839 19.55834
#> 4  14.25695 14.25712 14.25726 14.25739 14.25749 14.25757 14.25764 14.25770
#> 5  24.58978 24.59002 24.59022 24.59039 24.59054 24.59067 24.59077 24.59086
#> 6  22.75249 22.75255 22.75260 22.75265 22.75270 22.75275 22.75279 22.75283
#> 7  15.97087 15.97087 15.97088 15.97090 15.97091 15.97093 15.97094 15.97096
#> 8  15.84469 15.84468 15.84467 15.84467 15.84468 15.84469 15.84470 15.84471
#> 9  15.59232 15.59228 15.59225 15.59223 15.59222 15.59221 15.59221 15.59221
#> 10 10.72352 10.72341 10.72331 10.72322 10.72315 10.72308 10.72302 10.72296
#> 11 11.03560 11.03552 11.03545 11.03538 11.03533 11.03528 11.03523 11.03519
#> 12 11.57529 11.57529 11.57528 11.57528 11.57527 11.57526 11.57526 11.57525
#> 13 16.45627 16.45616 16.45606 16.45598 16.45591 16.45585 16.45581 16.45577
#> 14 14.44118 14.44146 14.44170 14.44191 14.44209 14.44224 14.44237 14.44248
#> 15 24.88482 24.88475 24.88469 24.88464 24.88460 24.88456 24.88452 24.88449
#> 16 26.57966 26.57971 26.57976 26.57979 26.57983 26.57985 26.57988 26.57989
#> 17 27.35660 27.35656 27.35653 27.35650 27.35647 27.35644 27.35641 27.35639
#> 18 16.71968 16.71939 16.71915 16.71895 16.71878 16.71865 16.71855 16.71847
#> 19 21.43918 21.43930 21.43940 21.43949 21.43956 21.43961 21.43966 21.43970
#> 20 14.66133 14.66133 14.66132 14.66132 14.66131 14.66131 14.66131 14.66131
#> 21 23.32133 23.32135 23.32136 23.32136 23.32136 23.32136 23.32135 23.32134
#>      0.0372   0.0347   0.0323   0.0302   0.0281   0.0262   0.0245   0.0228
#> 1  19.91354 19.91363 19.91372 20.24428 20.24428 20.24428 20.24427 20.24426
#> 2  24.57247 24.57253 24.57258 24.42162 24.42173 24.42184 24.42195 24.42206
#> 3  19.62369 19.62369 19.62369 19.56135 19.56139 19.56143 19.56147 19.56150
#> 4  14.64753 14.64787 14.64819 14.58604 14.58641 14.58677 14.58712 14.58744
#> 5  24.43148 24.43143 24.43140 24.50763 24.50751 24.50740 24.50729 24.50719
#> 6  22.96371 22.96371 22.96371 22.96440 22.96441 22.96442 22.96443 22.96443
#> 7  15.66435 15.66404 15.66376 15.58146 15.58114 15.58084 15.58056 15.58029
#> 8  15.79815 15.79799 15.79784 15.78600 15.78584 15.78568 15.78554 15.78539
#> 9  15.56342 15.56325 15.56309 15.55431 15.55414 15.55398 15.55383 15.55368
#> 10 10.79720 10.79739 10.79758 10.94943 10.94961 10.94979 10.94995 10.95010
#> 11 10.90757 10.90764 10.90769 10.94434 10.94440 10.94445 10.94451 10.94456
#> 12 11.34462 11.34459 11.34458 11.27761 11.27759 11.27757 11.27756 11.27755
#> 13 16.53220 16.53216 16.53211 16.65056 16.65050 16.65045 16.65039 16.65034
#> 14 14.52665 14.52678 14.52690 14.43781 14.43792 14.43803 14.43814 14.43825
#> 15 25.01832 25.01841 25.01848 25.05913 25.05923 25.05932 25.05940 25.05948
#> 16 26.39719 26.39707 26.39696 26.46512 26.46495 26.46478 26.46463 26.46448
#> 17 27.52487 27.52497 27.52508 27.45404 27.45418 27.45432 27.45446 27.45459
#> 18 16.74540 16.74514 16.74489 16.75708 16.75686 16.75665 16.75644 16.75624
#> 19 21.27135 21.27129 21.27123 21.36270 21.36258 21.36246 21.36235 21.36225
#> 20 14.77146 14.77153 14.77161 14.59477 14.59491 14.59504 14.59517 14.59530
#> 21 23.18482 23.18488 23.18493 23.04032 23.04040 23.04048 23.04055 23.04063
#>      0.0213   0.0198   0.0185   0.0173   0.0161   0.0150   0.0140   0.0131
#> 1  20.24425 20.24424 20.26710 20.26728 20.26746 20.26763 20.26780 20.26796
#> 2  24.42216 24.42225 24.39260 24.39253 24.39247 24.39241 24.39236 24.39231
#> 3  19.56153 19.56156 19.60834 19.60837 19.60841 19.60844 19.60846 19.60849
#> 4  14.58775 14.58805 14.53070 14.53085 14.53100 14.53115 14.53130 14.53145
#> 5  24.50710 24.50701 24.63841 24.63832 24.63822 24.63813 24.63803 24.63792
#> 6  22.96444 22.96444 22.90362 22.90369 22.90376 22.90383 22.90389 22.90396
#> 7  15.58003 15.57979 15.58345 15.58323 15.58301 15.58279 15.58258 15.58236
#> 8  15.78526 15.78513 15.75903 15.75898 15.75892 15.75887 15.75882 15.75876
#> 9  15.55354 15.55341 15.51054 15.51050 15.51046 15.51042 15.51038 15.51034
#> 10 10.95024 10.95038 10.93137 10.93156 10.93174 10.93193 10.93211 10.93228
#> 11 10.94460 10.94465 10.93062 10.93063 10.93065 10.93067 10.93069 10.93071
#> 12 11.27754 11.27753 11.27783 11.27773 11.27762 11.27752 11.27742 11.27733
#> 13 16.65028 16.65023 16.62995 16.63005 16.63014 16.63024 16.63033 16.63041
#> 14 14.43836 14.43846 14.49694 14.49690 14.49687 14.49684 14.49681 14.49678
#> 15 25.05955 25.05962 25.09789 25.09799 25.09809 25.09818 25.09827 25.09836
#> 16 26.46434 26.46421 26.28854 26.28840 26.28827 26.28815 26.28804 26.28792
#> 17 27.45471 27.45483 27.48387 27.48392 27.48396 27.48400 27.48404 27.48408
#> 18 16.75605 16.75586 16.83882 16.83891 16.83899 16.83907 16.83914 16.83921
#> 19 21.36215 21.36205 21.36259 21.36250 21.36241 21.36233 21.36224 21.36216
#> 20 14.59542 14.59553 14.59011 14.59010 14.59010 14.59010 14.59010 14.59010
#> 21 23.04071 23.04078 23.07768 23.07755 23.07744 23.07732 23.07721 23.07710
#>      0.0122   0.0114   0.0106   0.0099   0.0092   0.0086   0.0080   0.0075
#> 1  20.26812 20.26827 20.26841 20.26855 20.26869 20.26882 20.26894 20.26906
#> 2  24.39226 24.39221 24.39217 24.39213 24.39209 24.39206 24.39203 24.39200
#> 3  19.60852 19.60854 19.60857 19.60859 19.60861 19.60863 19.60865 19.60866
#> 4  14.53160 14.53175 14.53189 14.53204 14.53219 14.53234 14.53249 14.53263
#> 5  24.63782 24.63772 24.63761 24.63751 24.63740 24.63730 24.63719 24.63708
#> 6  22.90403 22.90409 22.90416 22.90423 22.90429 22.90436 22.90442 22.90449
#> 7  15.58216 15.58195 15.58175 15.58156 15.58136 15.58117 15.58098 15.58080
#> 8  15.75871 15.75866 15.75861 15.75856 15.75851 15.75846 15.75842 15.75837
#> 9  15.51030 15.51027 15.51023 15.51020 15.51016 15.51013 15.51010 15.51006
#> 10 10.93245 10.93262 10.93279 10.93295 10.93311 10.93327 10.93343 10.93358
#> 11 10.93073 10.93075 10.93077 10.93079 10.93081 10.93083 10.93085 10.93087
#> 12 11.27723 11.27714 11.27706 11.27697 11.27689 11.27681 11.27673 11.27665
#> 13 16.63050 16.63058 16.63066 16.63074 16.63082 16.63089 16.63096 16.63103
#> 14 14.49675 14.49671 14.49668 14.49665 14.49662 14.49659 14.49656 14.49653
#> 15 25.09845 25.09853 25.09862 25.09869 25.09877 25.09884 25.09892 25.09899
#> 16 26.28782 26.28772 26.28763 26.28754 26.28746 26.28738 26.28731 26.28724
#> 17 27.48412 27.48416 27.48420 27.48424 27.48427 27.48431 27.48435 27.48438
#> 18 16.83927 16.83932 16.83937 16.83941 16.83945 16.83949 16.83952 16.83955
#> 19 21.36208 21.36200 21.36192 21.36185 21.36177 21.36170 21.36162 21.36155
#> 20 14.59011 14.59011 14.59011 14.59012 14.59013 14.59014 14.59015 14.59016
#> 21 23.07699 23.07689 23.07679 23.07669 23.07659 23.07650 23.07641 23.07632
#>      0.0070   0.0065   0.0061   0.0057   0.0053   0.0049
#> 1  20.26918 20.26929 20.26940 20.26950 20.26960 20.26970
#> 2  24.39197 24.39195 24.39192 24.39191 24.39189 24.39187
#> 3  19.60868 19.60869 19.60871 19.60872 19.60873 19.60874
#> 4  14.53278 14.53292 14.53306 14.53321 14.53335 14.53349
#> 5  24.63697 24.63686 24.63676 24.63665 24.63654 24.63643
#> 6  22.90455 22.90461 22.90467 22.90474 22.90480 22.90486
#> 7  15.58062 15.58045 15.58027 15.58010 15.57994 15.57978
#> 8  15.75833 15.75828 15.75824 15.75820 15.75816 15.75812
#> 9  15.51003 15.51000 15.50997 15.50994 15.50992 15.50989
#> 10 10.93372 10.93387 10.93401 10.93415 10.93428 10.93442
#> 11 10.93089 10.93091 10.93093 10.93095 10.93097 10.93099
#> 12 11.27658 11.27650 11.27643 11.27637 11.27630 11.27624
#> 13 16.63109 16.63116 16.63122 16.63128 16.63134 16.63140
#> 14 14.49650 14.49648 14.49645 14.49642 14.49639 14.49637
#> 15 25.09905 25.09912 25.09918 25.09924 25.09930 25.09935
#> 16 26.28717 26.28711 26.28706 26.28701 26.28696 26.28692
#> 17 27.48442 27.48445 27.48448 27.48451 27.48455 27.48458
#> 18 16.83957 16.83959 16.83961 16.83962 16.83963 16.83963
#> 19 21.36148 21.36141 21.36135 21.36128 21.36121 21.36115
#> 20 14.59017 14.59018 14.59019 14.59021 14.59022 14.59024
#> 21 23.07624 23.07616 23.07608 23.07600 23.07593 23.07585
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 22.8 18.1 14.3 24.4 22.8 16.4 17.3 15.2 10.4 10.4 14.7 15.5 13.3 27.3
#> [16] 26.0 30.4 15.8 19.7 15.0 21.4
#> 
#> $X
#>               am        carb        cyl       disp        drat       gear
#>  [1,]  1.1547005  0.51323935 -0.2102353 -0.7174329  0.76720598  0.2936101
#>  [2,]  1.1547005 -1.28309838 -1.3139708 -1.1385675  0.66574927  0.2936101
#>  [3,] -0.8660254 -1.28309838 -0.2102353 -0.1910146 -1.54600702 -0.9395524
#>  [4,] -0.8660254  0.51323935  0.8935002  0.9023156 -0.63289662 -0.9395524
#>  [5,] -0.8660254 -0.68431914 -1.3139708 -0.8251462  0.34108780  0.2936101
#>  [6,] -0.8660254 -0.68431914 -1.3139708 -0.8729288  0.80778867  0.2936101
#>  [7,] -0.8660254 -0.08553989  0.8935002  0.2204015 -0.91697541 -0.9395524
#>  [8,] -0.8660254 -0.08553989  0.8935002  0.2204015 -0.91697541 -0.9395524
#>  [9,] -0.8660254 -0.08553989  0.8935002  0.2204015 -0.91697541 -0.9395524
#> [10,] -0.8660254  0.51323935  0.8935002  1.8093748 -1.20105420 -0.9395524
#> [11,] -0.8660254  0.51323935  0.8935002  1.7121899 -1.05901481 -0.9395524
#> [12,] -0.8660254  0.51323935  0.8935002  1.5502151 -0.59231394 -0.9395524
#> [13,] -0.8660254 -0.68431914  0.8935002  0.5621684 -1.54600702 -0.9395524
#> [14,] -0.8660254  0.51323935  0.8935002  0.8213282  0.42225317 -0.9395524
#> [15,]  1.1547005 -1.28309838 -1.3139708 -1.3734311  1.13245014  0.2936101
#> [16,]  1.1547005 -0.68431914 -1.3139708 -1.0389530  1.84264711  1.5267726
#> [17,]  1.1547005 -0.68431914 -1.3139708 -1.2430413  0.50341853  1.5267726
#> [18,]  1.1547005  0.51323935  0.8935002  0.8294270  1.41652893  1.5267726
#> [19,]  1.1547005  1.71079785 -0.2102353 -0.8389140  0.19904840  1.5267726
#> [20,]  1.1547005  2.90835634  0.8935002  0.4244898  0.03671767  1.5267726
#> [21,]  1.1547005 -0.68431914 -1.3139708 -1.0332839  1.19332417  0.2936101
#>               hp        qsec         vs          wt
#>  [1,] -0.7513242 -0.55480559 -0.7071068 -0.74555226
#>  [2,] -0.9870208  0.55505140  1.4142136 -1.03680167
#>  [3,] -0.8206467  1.38615361  1.4142136  0.06994609
#>  [4,]  1.1203842 -0.87485737 -0.7071068  0.17673754
#>  [5,] -1.4168205  1.27258685  1.4142136 -0.19217838
#>  [6,] -0.9592918  2.76960326  1.4142136 -0.23101164
#>  [7,]  0.2191912 -0.06956579 -0.7071068  0.66215322
#>  [8,]  0.2191912  0.03367672 -0.7071068  0.33207056
#>  [9,]  0.2191912  0.24016175 -0.7071068  0.38061213
#> [10,]  0.5658039  0.22983749 -0.7071068  1.80773424
#> [11,]  0.7044490  0.14724349 -0.7071068  1.97665889
#> [12,]  0.9124166 -0.05924154 -0.7071068  1.89996322
#> [13,] -0.1967439 -0.34315844 -0.7071068  0.12819597
#> [14,]  1.1203842 -1.09682877 -0.7071068  0.43886201
#> [15,] -1.3613625  0.70475304  1.4142136 -1.41057175
#> [16,] -1.0147498 -0.43091457 -0.7071068 -1.21155132
#> [17,] -0.7097307 -0.32767206  1.4142136 -1.82026258
#> [18,]  1.3838098 -1.56658219 -0.7071068 -0.21159501
#> [19,]  0.1498687 -1.05036964 -0.7071068 -0.59992755
#> [20,]  2.3681897 -1.51496094 -0.7071068  0.17673754
#> [21,] -0.7651887  0.54988928  1.4142136 -0.59021924
#> attr(,"center")
#>  [1]   0.4285714   3.1428571   6.3809524 248.5857143   3.5219048   3.7619048
#>  [7] 164.1904762  17.5347619   0.3333333   3.3879524
#> attr(,"scale")
#>  [1]   0.4948717   1.6700646   1.8120283 123.4759560   0.4928210   0.8109232
#>  [7]  72.1266208   1.9371865   0.4714045   1.0300450
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 43
#> 
#> $lambda.min
#> [1] 0.2623907
#> 
#> $null.dev
#> [1] 28.49134
#> 
#> $Bias
#> [1] 1.175893
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 17.54362