Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 41.511154 38.752542 35.425860 30.905841 27.121008 24.084966 21.682942
#>   [8] 19.816277 18.400245 17.362164 16.639765 16.179767 15.936648 15.871581
#>  [15] 15.951509 16.146406 16.138255 16.007366 15.849207 15.553326 15.193503
#>  [22] 14.769797 14.302078 13.917119 13.605591 12.030340 11.833433 11.681234
#>  [29] 11.918702 12.398874 12.707526 12.520758 11.611353 11.559686 11.316067
#>  [36] 11.197547 11.024386 11.015037 11.014907 11.014811 10.852116 11.061880
#>  [43] 10.009786  9.769614 11.691571 12.103835 12.773901 17.322863 17.323270
#>  [50] 16.227647 16.227830 15.919106 16.327906 16.328025 19.207580 19.208279
#>  [57] 19.208955 32.841274 32.686019 35.513625 35.514809 35.734723 35.849261
#>  [64] 35.192796 36.273080 36.274115 36.275104 35.908160 35.998082 35.999158
#>  [71] 36.048831 36.049853 35.990338 35.991101 35.998492 36.192768 36.164323
#>  [78] 36.165540 36.166739 36.167921 36.169086 36.170235 36.171369 36.172487
#>  [85] 36.173591 36.174681 36.175758 35.955058 35.316741 35.315823 35.312783
#>  [92] 35.309789 35.306840 35.303934 35.301071 35.298250 35.295471 35.292732
#>  [99] 35.290033 35.287372
#> 
#> $cvse
#>   [1] 10.726637 10.406183 10.031432  8.848310  7.812634  6.952182  6.232805
#>   [8]  5.629543  5.125624  4.711159  4.381501  4.135266  3.972064  3.890274
#>  [15]  3.885439  3.950032  4.024742  3.958489  3.925945  3.943204  3.933595
#>  [22]  3.875298  3.803368  3.734808  3.671129  3.456655  3.399617  3.351185
#>  [29]  3.294924  3.230795  3.200620  3.172150  3.322477  3.289429  3.260288
#>  [36]  3.285631  3.301507  3.297226  3.297220  3.297216  3.210718  3.077254
#>  [43]  2.842690  2.638192  3.229312  3.190878  3.238402  6.630078  6.630354
#>  [50]  6.590016  6.590223  6.612003  6.708898  6.709028  6.792101  6.792285
#>  [57]  6.792453 18.029764 18.029428 18.118318 18.119871 18.101711 18.089281
#>  [64] 18.114284 18.040241 18.041704 18.043152 18.044743 18.039398 18.040782
#>  [71] 18.037406 18.038758 18.043152 18.044489 18.046178 18.040524 18.044592
#>  [78] 18.045845 18.047083 18.048305 18.049512 18.050705 18.051883 18.053046
#>  [85] 18.054195 18.055330 18.056451 18.068647 17.525271 17.522702 17.520475
#>  [92] 17.518284 17.516129 17.514008 17.511921 17.509867 17.507846 17.505857
#>  [99] 17.503899 17.501972
#> 
#> $fold
#>  [1]  1  8  3  7  9  6  7  8  6  1  5  5  2  1  9 10 10  3  4  2  4
#> 
#> $lambda
#>   [1] 5.506553220 5.135429963 4.789319171 4.466535088 4.165505572 3.884764438
#>   [7] 3.622944317 3.378769995 3.151052206 2.938681834 2.740624513 2.555915594
#>  [13] 2.383655437 2.223005038 2.073181939 1.933456415 1.803147924 1.681621789
#>  [19] 1.568286108 1.462588873 1.364015278 1.272085214 1.186350928 1.106394846
#>  [25] 1.031827536 0.962285814 0.897430971 0.836947127 0.780539692 0.727933929
#>  [31] 0.678873618 0.633119809 0.590449652 0.550655322 0.513542996 0.478931917
#>  [37] 0.446653509 0.416550557 0.388476444 0.362294432 0.337877000 0.315105221
#>  [43] 0.293868183 0.274062451 0.255591559 0.238365544 0.222300504 0.207318194
#>  [49] 0.193345642 0.180314792 0.168162179 0.156828611 0.146258887 0.136401528
#>  [55] 0.127208521 0.118635093 0.110639485 0.103182754 0.096228581 0.089743097
#>  [61] 0.083694713 0.078053969 0.072793393 0.067887362 0.063311981 0.059044965
#>  [67] 0.055065532 0.051354300 0.047893192 0.044665351 0.041655056 0.038847644
#>  [73] 0.036229443 0.033787700 0.031510522 0.029386818 0.027406245 0.025559156
#>  [79] 0.023836554 0.022230050 0.020731819 0.019334564 0.018031479 0.016816218
#>  [85] 0.015682861 0.014625889 0.013640153 0.012720852 0.011863509 0.011063948
#>  [91] 0.010318275 0.009622858 0.008974310 0.008369471 0.007805397 0.007279339
#>  [97] 0.006788736 0.006331198 0.005904497 0.005506553
#> 
#> $fit
#> $beta
#>               5.5066     5.1354    4.7893    4.4665    4.1655    3.8848
#> (Intercept) 21.30952 22.9592087 24.497710 25.932522 27.270632 28.518558
#> am           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000 -0.5442623 -1.051843 -1.525215 -1.966683 -2.378397
#>                3.6229    3.3788    3.1511    2.9387    2.7406   2.5559
#> (Intercept) 29.682378 30.767761 31.779992 32.724003 33.604390 34.42544
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> wt          -2.762363 -3.120452 -3.454406 -3.765853 -4.056309 -4.32719
#>                2.3837    2.2230    2.0732    1.9335    1.8031    1.6816
#> (Intercept) 35.191158 35.905267 36.571248 37.192343 37.627692 37.627692
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -4.579814 -4.815412 -5.035132 -5.240043 -5.383673 -5.383673
#>                  1.5683     1.4626     1.3640     1.2721     1.1864     1.1064
#> (Intercept) 36.89477782 35.0833728 33.3944976 31.8194469 30.3505495 28.9806507
#> am           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.03788363  0.1315134  0.2188098  0.3002226  0.3761485  0.4469573
#> vs           0.00000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.36615345 -5.3228547 -5.2824848 -5.2448357 -5.2097240 -5.1769788
#>                 1.0318     0.9623     0.8974    0.8369     0.7805     0.7279
#> (Intercept) 27.7030786 26.5116106 25.4004435 24.364165 23.3977289 22.4964270
#> am           0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> qsec         0.5129938  0.5745796  0.6320148  0.685579  0.7355332  0.7821206
#> vs           0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> wt          -5.1464404 -5.1179602 -5.0913996 -5.066629 -5.0435278 -5.0219836
#>                 0.6789    0.6331    0.5904    0.5507    0.5135     0.4789
#> (Intercept) 21.6558699 15.212332 12.179573  9.350304  9.036187 16.3528599
#> am           0.0000000  1.407730  2.327530  3.185602  3.280900  4.1339911
#> carb         0.0000000  0.000000  0.000000  0.000000  0.000000 -0.8085199
#> cyl          0.0000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> disp         0.0000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> drat         0.0000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> gear         0.0000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> hp           0.0000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> qsec         0.8255682  1.047144  1.131061  1.209349  1.218038  0.7847179
#> vs           0.0000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> wt          -5.0018914 -4.431121 -4.086328 -3.764673 -3.728952 -2.9500618
#>                 0.4467     0.4166     0.3885     0.3623     0.3379     0.3151
#> (Intercept) 16.3518812 16.3511500 16.3506128 16.3502230 16.3499425 16.3497420
#> am           4.1348121  4.1353940  4.1358051  4.1360949  4.1362988  4.1364421
#> carb        -0.8086976 -0.8088206 -0.8089059 -0.8089652 -0.8090064 -0.8090351
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.7847062  0.7846998  0.7846963  0.7846944  0.7846934  0.7846929
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.9496413 -2.9493449 -2.9491363 -2.9489898 -2.9488870 -2.9488149
#>                 0.2939     0.2741     0.2556     0.2384     0.2223     0.2073
#> (Intercept) 16.3495994 16.3494984 16.3494271 16.3493768 16.3493414 16.3493166
#> am           4.1365427  4.1366132  4.1366626  4.1366972  4.1367214  4.1367383
#> carb        -0.8090551 -0.8090690 -0.8090787 -0.8090855 -0.8090902 -0.8090935
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.7846926  0.7846924  0.7846923  0.7846923  0.7846923  0.7846923
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.9487643 -2.9487290 -2.9487042 -2.9486869 -2.9486748 -2.9486663
#>                 0.1933     0.1803     0.1682     0.1568     0.1463     0.1364
#> (Intercept) -27.771207 -27.775510 -27.779506 -27.783215 -27.786660 -27.789858
#> am            3.549039   3.548761   3.548501   3.548260   3.548035   3.547827
#> carb         -2.349459  -2.349774  -2.350066  -2.350338  -2.350590  -2.350824
#> cyl           2.574119   2.574380   2.574621   2.574846   2.575054   2.575247
#> disp          0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> drat          2.802285   2.802670   2.803028   2.803360   2.803669   2.803956
#> gear          3.976955   3.977658   3.978312   3.978918   3.979482   3.980006
#> hp            0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> qsec          1.235491   1.235448   1.235408   1.235370   1.235335   1.235303
#> vs            0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> wt           -2.787626  -2.787443  -2.787274  -2.787116  -2.786970  -2.786834
#>                 0.1272     0.1186     0.1106     0.1032     0.0962     0.0897
#> (Intercept) -27.792828 -27.795586 -27.798147 -27.800525 -27.802733 -27.804784
#> am            3.547633   3.547452   3.547284   3.547128   3.546983   3.546848
#> carb         -2.351042  -2.351244  -2.351432  -2.351607  -2.351769  -2.351920
#> cyl           2.575427   2.575593   2.575748   2.575892   2.576025   2.576149
#> disp          0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> drat          2.804222   2.804470   2.804700   2.804914   2.805113   2.805297
#> gear          3.980492   3.980944   3.981364   3.981754   3.982117   3.982454
#> hp            0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> qsec          1.235273   1.235245   1.235218   1.235194   1.235171   1.235150
#> vs            0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> wt           -2.786707  -2.786590  -2.786480  -2.786379  -2.786284  -2.786197
#>                 0.0837     0.0781     0.0728     0.0679     0.0633     0.0590
#> (Intercept) -27.806688 -27.808456 -27.810099 -27.811624 -27.813040 -27.814356
#> am            3.546722   3.546605   3.546497   3.546396   3.546302   3.546215
#> carb         -2.352060  -2.352190  -2.352310  -2.352423  -2.352527  -2.352624
#> cyl           2.576264   2.576371   2.576470   2.576562   2.576648   2.576727
#> disp          0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> drat          2.805469   2.805628   2.805777   2.805914   2.806042   2.806161
#> gear          3.982767   3.983057   3.983327   3.983578   3.983811   3.984028
#> hp            0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> qsec          1.235130   1.235112   1.235095   1.235079   1.235064   1.235051
#> vs            0.000000   0.000000   0.000000   0.000000   0.000000   0.000000
#> wt           -2.786115  -2.786039  -2.785969  -2.785903  -2.785842  -2.785786
#>                 0.0551        0.0514        0.0479        0.0447        0.0417
#> (Intercept) -27.815578 -26.438083906 -26.434223226 -26.430494507 -26.426893287
#> am            3.546134   3.675777414   3.676032768   3.676279635   3.676518288
#> carb         -2.352714  -1.889618030  -1.889284366  -1.888962041  -1.888650678
#> cyl           2.576801   2.119159054   2.118808155   2.118469266   2.118141979
#> disp          0.000000   0.015784082   0.015788529   0.015792823   0.015796969
#> drat          2.806271   2.601073942   2.600692941   2.600324867   2.599969291
#> gear          3.984229   3.850646173   3.850095334   3.849563050   3.849048716
#> hp            0.000000  -0.009298851  -0.009300345  -0.009301787  -0.009303178
#> qsec          1.235038   1.368830164   1.368895208   1.368958091   1.369018881
#> vs            0.000000   0.000000000   0.000000000   0.000000000   0.000000000
#> wt           -2.785733  -3.837447098  -3.837878274  -3.838294752  -3.838697028
#>                   0.0388        0.0362        0.0338        0.0315
#> (Intercept) -26.42341524 -26.420056199 -26.416812102 -26.413679042
#> am            3.67674899   3.676971996   3.677187555   3.677395905
#> carb         -1.88834991  -1.888059388  -1.887778765  -1.887507710
#> cyl           2.11782590   2.117520643   2.117225841   2.116941137
#> disp          0.01580097   0.015804837   0.015808568   0.015812171
#> drat          2.59962580   2.599293979   2.598973450   2.598663832
#> gear          3.84855174   3.848071553   3.847607600   3.847159345
#> hp           -0.00930452  -0.009305814  -0.009307062  -0.009308266
#> qsec          1.36907764   1.369134445   1.369189347   1.369242410
#> vs            0.00000000   0.000000000   0.000000000   0.000000000
#> wt           -3.83908558  -3.839460865  -3.839823334  -3.840173417
#>                    0.0294        0.0274        0.0256       0.0238
#> (Intercept) -26.410653231 -26.407731010 -26.404908841 -26.40218330
#> am            3.677597281   3.677791907   3.677980001   3.67816178
#> carb         -1.887245903  -1.886993031  -1.886748796  -1.88651290
#> cyl           2.116666186   2.116400653   2.116144218   2.11589657
#> disp          0.015815650   0.015819008   0.015822251   0.01582538
#> drat          2.598364758   2.598075874   2.597796838   2.59752732
#> gear          3.846726266   3.846307862   3.845903643   3.84551314
#> hp           -0.009309427  -0.009310547  -0.009311628  -0.00931267
#> qsec          1.369293694   1.369343254   1.369391147   1.36943743
#> vs            0.000000000   0.000000000   0.000000000   0.00000000
#> wt           -3.840511531  -3.840838081  -3.841153455  -3.84145803
#>                    0.0222        0.0207        0.0193        0.0180
#> (Intercept) -26.399551095 -26.397009021 -26.394553997 -26.392183044
#> am            3.678337437   3.678507182   3.678671205   3.678829692
#> carb         -1.886285076  -1.886065037  -1.885852526  -1.885647287
#> cyl           2.115657406   2.115426439   2.115203387   2.114987979
#> disp          0.015828404   0.015831323   0.015834141   0.015836861
#> drat          2.597266996   2.597015559   2.596772709   2.596538155
#> gear          3.845135888   3.844771452   3.844419401   3.844079320
#> hp           -0.009313676  -0.009314646  -0.009315581  -0.009316484
#> qsec          1.369482142   1.369525348   1.369567092   1.369607421
#> vs            0.000000000   0.000000000   0.000000000   0.000000000
#> wt           -3.841752176  -3.842036239  -3.842310563  -3.842575478
#>                    0.0168        0.0157        0.0146        0.0136
#> (Intercept) -26.389893283 -26.387681935 -26.385546316 -26.383483832
#> am            3.678982825   3.679130778   3.679273723   3.679411823
#> carb         -1.885449073  -1.885257647  -1.885072778  -1.884894244
#> cyl           2.114779955   2.114579061   2.114385055   2.114197700
#> disp          0.015839488   0.015842024   0.015844472   0.015846836
#> drat          2.596311616   2.596092823   2.595881511   2.595677429
#> gear          3.843750808   3.843433477   3.843126951   3.842830867
#> hp           -0.009317354  -0.009318194  -0.009319005  -0.009319786
#> qsec          1.369646381   1.369684018   1.369720374   1.369755492
#> vs            0.000000000   0.000000000   0.000000000   0.000000000
#> wt           -3.842831302  -3.843078343  -3.843316899  -3.843547259
#>                   0.0127        0.0119       0.0111        0.0103        0.0096
#> (Intercept) -26.38149198 -26.379568345 -26.37771059 -26.375916461 -26.374183783
#> am            3.67954524   3.679674123   3.67979863   3.679918897   3.680035072
#> carb         -1.88472183  -1.884555325  -1.88439453  -1.884239253  -1.884089303
#> cyl           2.11401677   2.113842042   2.11367331   2.113510362   2.113353006
#> disp          0.01584912   0.015851321   0.01585345   0.015855502   0.015857484
#> drat          2.59548033   2.595289981   2.59510615   2.594928613   2.594757161
#> gear          3.84254487   3.842268632   3.84200181   3.841744097   3.841495179
#> hp           -0.00932054  -0.009321268  -0.00932197  -0.009322647  -0.009323301
#> qsec          1.36978941   1.369822175   1.36985382   1.369884378   1.369913892
#> vs            0.00000000   0.000000000   0.00000000   0.000000000   0.000000000
#> wt           -3.84376970  -3.843984493  -3.84419190  -3.844392164  -3.844585536
#>                    0.0090        0.0084        0.0078        0.0073
#> (Intercept) -26.372510454 -26.370894444 -26.369333795 -26.367826614
#> am            3.680147288   3.680255677   3.680360367   3.680461481
#> carb         -1.883944499  -1.883804666  -1.883669635  -1.883539241
#> cyl           2.113201048   2.113054305   2.112912598   2.112775755
#> disp          0.015859398   0.015861246   0.015863030   0.015864752
#> drat          2.594591586   2.594431687   2.594277271   2.594128153
#> gear          3.841254762   3.841022558   3.840798290   3.840581688
#> hp           -0.009323931  -0.009324539  -0.009325126  -0.009325693
#> qsec          1.369942395   1.369969920   1.369996499   1.370022166
#> vs            0.000000000   0.000000000   0.000000000   0.000000000
#> wt           -3.844772249  -3.844952531  -3.845126600  -3.845294668
#>                    0.0068        0.0063        0.0059        0.0055
#> (Intercept) -26.366371075 -26.364965412 -26.363607924 -26.362296964
#> am            3.680559139   3.680653456   3.680744545   3.680832515
#> carb         -1.883413327  -1.883291740  -1.883174331  -1.883060958
#> cyl           2.112643610   2.112516002   2.112392777   2.112273785
#> disp          0.015866415   0.015868020   0.015869569   0.015871065
#> drat          2.593984151   2.593845092   2.593710807   2.593581133
#> gear          3.840372494   3.840170457   3.839975334   3.839786891
#> hp           -0.009326239  -0.009326766  -0.009327275  -0.009327766
#> qsec          1.370046950   1.370070881   1.370093988   1.370116299
#> vs            0.000000000   0.000000000   0.000000000   0.000000000
#> wt           -3.845456942  -3.845613618  -3.845764890  -3.845910940
#> 
#> $iter
#>   [1]   0   3   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   4   4   4   4   4   4   4   4   4   4   4   4   4  47  35  35   8  74
#>  [37]   1   1   1   1   1   1   1   1   1   1   1   1 189   1   1   1   1   1
#>  [55]   1   1   1   1   1   1   1   1   1   1   1   1   1 240   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.506553220 5.135429963 4.789319171 4.466535088 4.165505572 3.884764438
#>   [7] 3.622944317 3.378769995 3.151052206 2.938681834 2.740624513 2.555915594
#>  [13] 2.383655437 2.223005038 2.073181939 1.933456415 1.803147924 1.681621789
#>  [19] 1.568286108 1.462588873 1.364015278 1.272085214 1.186350928 1.106394846
#>  [25] 1.031827536 0.962285814 0.897430971 0.836947127 0.780539692 0.727933929
#>  [31] 0.678873618 0.633119809 0.590449652 0.550655322 0.513542996 0.478931917
#>  [37] 0.446653509 0.416550557 0.388476444 0.362294432 0.337877000 0.315105221
#>  [43] 0.293868183 0.274062451 0.255591559 0.238365544 0.222300504 0.207318194
#>  [49] 0.193345642 0.180314792 0.168162179 0.156828611 0.146258887 0.136401528
#>  [55] 0.127208521 0.118635093 0.110639485 0.103182754 0.096228581 0.089743097
#>  [61] 0.083694713 0.078053969 0.072793393 0.067887362 0.063311981 0.059044965
#>  [67] 0.055065532 0.051354300 0.047893192 0.044665351 0.041655056 0.038847644
#>  [73] 0.036229443 0.033787700 0.031510522 0.029386818 0.027406245 0.025559156
#>  [79] 0.023836554 0.022230050 0.020731819 0.019334564 0.018031479 0.016816218
#>  [85] 0.015682861 0.014625889 0.013640153 0.012720852 0.011863509 0.011063948
#>  [91] 0.010318275 0.009622858 0.008974310 0.008369471 0.007805397 0.007279339
#>  [97] 0.006788736 0.006331198 0.005904497 0.005506553
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 31
#> 
#> $loss
#>   [1] 840.95810 718.71853 616.44695 531.26982 460.70625 402.61562 355.15213
#>   [8] 316.72530 285.96571 261.69528 242.90146 228.71488 218.38990 211.28782
#>  [15] 206.86228 204.64662 204.19340 204.19340 199.49114 188.71873 179.76432
#>  [22] 172.36071 166.28000 161.32570 157.32857 154.14292 151.64345 149.72249
#>  [29] 148.28766 147.25973 146.57087 129.75974 123.94707 121.93066 121.90958
#>  [36] 106.52379 106.52378 106.52377 106.52377 106.52377 106.52377 106.52377
#>  [43] 106.52377 106.52377 106.52377 106.52377 106.52377 106.52377  76.48859
#>  [50]  76.48857  76.48856  76.48854  76.48853  76.48852  76.48851  76.48851
#>  [57]  76.48850  76.48850  76.48849  76.48849  76.48848  76.48848  76.48848
#>  [64]  76.48847  76.48847  76.48847  76.48847  74.17625  74.17622  74.17620
#>  [71]  74.17617  74.17614  74.17612  74.17610  74.17608  74.17606  74.17604
#>  [78]  74.17603  74.17601  74.17600  74.17598  74.17597  74.17596  74.17595
#>  [85]  74.17594  74.17593  74.17592  74.17591  74.17591  74.17590  74.17589
#>  [92]  74.17589  74.17588  74.17587  74.17587  74.17587  74.17586  74.17586
#>  [99]  74.17585  74.17585
#> 
#> $linear.predictors
#>      5.5066   5.1354   4.7893   4.4665   4.1655   3.8848   3.6229   3.3788
#> 1  21.30952 21.53324 21.74188 21.93646 22.11792 22.28716 22.44499 22.59218
#> 2  21.30952 21.39445 21.47366 21.54753 21.61642 21.68067 21.74058 21.79646
#> 3  21.30952 21.69652 22.05743 22.39402 22.70793 23.00068 23.27370 23.52831
#> 4  21.30952 21.20941 21.11603 21.02896 20.94775 20.87201 20.80138 20.73551
#> 5  21.30952 21.08695 20.87937 20.68578 20.50524 20.33687 20.17985 20.03341
#> 6  21.30952 21.01619 20.74263 20.48751 20.24958 20.02768 19.82074 19.62775
#> 7  21.30952 21.24478 21.18440 21.12810 21.07558 21.02661 20.98093 20.93834
#> 8  21.30952 21.08695 20.87937 20.68578 20.50524 20.33687 20.17985 20.03341
#> 9  21.30952 20.92911 20.57434 20.24347 19.93491 19.64714 19.37876 19.12848
#> 10 21.30952 20.00713 18.79251 17.65976 16.60335 15.61813 14.69932 13.84243
#> 11 21.30952 20.05013 18.87561 17.78025 16.75871 15.80603 14.91755 14.08895
#> 12 21.30952 21.76183 22.18366 22.57705 22.94393 23.28608 23.60518 23.90277
#> 13 21.30952 22.08023 22.79898 23.46930 24.09444 24.67745 25.22116 25.72823
#> 14 21.30952 21.96049 22.56758 23.13375 23.66177 24.15420 24.61344 25.04173
#> 15 21.30952 21.61760 21.90492 22.17287 22.42276 22.65581 22.87315 23.07585
#> 16 21.30952 20.86924 20.45863 20.07570 19.71857 19.38551 19.07490 18.78523
#> 17 21.30952 21.90606 22.46239 22.98123 23.46510 23.91636 24.33721 24.72969
#> 18 21.30952 22.13574 22.90627 23.62487 24.29504 24.92004 25.50292 26.04652
#> 19 21.30952 21.45160 21.58410 21.70768 21.82292 21.93040 22.03063 22.12411
#> 20 21.30952 21.01619 20.74263 20.48751 20.24958 20.02768 19.82074 19.62775
#> 21 21.30952 21.44616 21.57359 21.69243 21.80325 21.90661 22.00301 22.09291
#>      3.1511   2.9387   2.7406   2.5559   2.3837    2.2230    2.0732    1.9335
#> 1  22.72945 22.85747 22.97686 23.08820 23.19205 23.288887 23.379203 23.463431
#> 2  21.84858 21.89718 21.94250 21.98477 22.02419 22.060957 22.095244 22.127220
#> 3  23.76577 23.98722 24.19375 24.38636 24.56599 24.733511 24.889742 25.035444
#> 4  20.67408 20.61679 20.56336 20.51353 20.46706 20.423717 20.383299 20.345605
#> 5  19.89684 19.76947 19.65069 19.53991 19.43660 19.340250 19.250395 19.166596
#> 6  19.44776 19.27991 19.12337 18.97737 18.84122 18.714246 18.595828 18.485390
#> 7  20.89861 20.86157 20.82702 20.79479 20.76474 20.736719 20.710583 20.686208
#> 8  19.89684 19.76947 19.65069 19.53991 19.43660 19.340250 19.250395 19.166596
#> 9  18.89506 18.67737 18.47436 18.28502 18.10845 17.943780 17.790207 17.646983
#> 10 13.04330 12.29802 11.60297 10.95476 10.35025  9.786472  9.260694  8.770351
#> 11 13.31619 12.59552 11.92342 11.29661 10.71205 10.166890  9.658469  9.184314
#> 12 24.18030 24.43913 24.68051 24.90562 25.11557 25.311361 25.493958 25.664249
#> 13 26.20113 26.64215 27.05345 27.43703 27.79476 28.128377 28.439510 28.729674
#> 14 25.44116 25.81366 26.16106 26.48505 26.78720 27.068986 27.331781 27.576865
#> 15 23.26488 23.44118 23.60559 23.75892 23.90192 24.035276 24.159648 24.275638
#> 16 18.51507 18.26313 18.02816 17.80903 17.60467 17.414085 17.236342 17.070579
#> 17 25.09572 25.43708 25.75543 26.05233 26.32922 26.587445 26.828268 27.052860
#> 18 26.55348 27.02627 27.46719 27.87840 28.26190 28.619549 28.953093 29.264158
#> 19 22.21129 22.29259 22.36841 22.43913 22.50507 22.566576 22.623933 22.677425
#> 20 19.44776 19.27991 19.12337 18.97737 18.84122 18.714246 18.595828 18.485390
#> 21 22.17674 22.25493 22.32785 22.39585 22.45928 22.518422 22.573582 22.625024
#>       1.8031    1.6816    1.5683    1.4626    1.3640    1.2721    1.1864
#> 1  23.522470 23.522470 23.459020 23.302204 23.155996 23.019642 22.892477
#> 2  22.149633 22.149633 22.111866 22.018524 21.931496 21.850333 21.774641
#> 3  25.137571 25.137571 25.150316 25.181815 25.211183 25.238571 25.264114
#> 4  20.319184 20.319184 20.379052 20.527016 20.664971 20.793628 20.913614
#> 5  19.107858 19.107858 19.079989 19.011111 18.946892 18.887001 18.831147
#> 6  18.407981 18.407981 18.337687 18.163954 18.001974 17.850910 17.710028
#> 7  20.669123 20.669123 20.858929 21.328038 21.765414 22.173313 22.553720
#> 8  19.107858 19.107858 19.151210 19.258356 19.358255 19.451420 19.538306
#> 9  17.546593 17.546593 17.545777 17.543761 17.541881 17.540128 17.538493
#> 10  8.426652  8.426652  8.463848  8.555778  8.641490  8.721425  8.795973
#> 11  8.851962  8.851962  8.872620  8.923678  8.971283  9.015678  9.057082
#> 12 25.783612 25.783612 25.826834 25.933659 26.033257 26.126143 26.212769
#> 13 28.933061 28.933061 28.930045 28.922591 28.915642 28.909160 28.903116
#> 14 27.748653 27.748653 27.801770 27.933052 28.055452 28.169604 28.276062
#> 15 24.356939 24.356939 24.425261 24.594120 24.751556 24.898382 25.035312
#> 16 16.954389 16.954389 16.872535 16.670233 16.481615 16.305709 16.141658
#> 17 27.210285 27.210285 27.227271 27.269253 27.308394 27.344898 27.378941
#> 18 29.482195 29.482195 29.416021 29.252470 29.099983 28.957773 28.825147
#> 19 22.714919 22.714919 22.617729 22.377523 22.153566 21.944703 21.749916
#> 20 18.407981 18.407981 18.290711 18.000878 17.730649 17.478634 17.243603
#> 21 22.661082 22.661082 22.681507 22.731986 22.779052 22.822945 22.863879
#>       1.1064    1.0318    0.9623    0.8974    0.8369    0.7805    0.7279
#> 1  22.773883 22.663282 22.560135 22.463940 22.374228 22.290562 22.212535
#> 2  21.704050 21.638216 21.576820 21.519561 21.466162 21.416361 21.369917
#> 3  25.287935 25.310151 25.330870 25.350192 25.368212 25.385017 25.400690
#> 4  21.025514 21.129872 21.227196 21.317961 21.402609 21.481552 21.555175
#> 5  18.779057 18.730478 18.685172 18.642921 18.603516 18.566768 18.532496
#> 6  17.578640 17.456108 17.341834 17.235261 17.135871 17.043180 16.956736
#> 7  22.908490 23.239349 23.547909 23.835673 24.104043 24.354326 24.587741
#> 8  19.619337 19.694906 19.765382 19.831108 19.892405 19.949570 20.002883
#> 9  17.536968 17.535546 17.534220 17.532983 17.531830 17.530754 17.529751
#> 10  8.865497  8.930335  8.990803  9.047196  9.099788  9.148835  9.194577
#> 11  9.095695  9.131706  9.165290  9.196610  9.225820  9.253061  9.278466
#> 12 26.293556 26.368899 26.439163 26.504692 26.565805 26.622799 26.675952
#> 13 28.897479 28.892222 28.887319 28.882747 28.878483 28.874506 28.870797
#> 14 28.375345 28.467937 28.554288 28.634820 28.709924 28.779966 28.845287
#> 15 25.163014 25.282108 25.393177 25.496759 25.593361 25.683452 25.767471
#> 16 15.988664 15.845982 15.712915 15.588817 15.473083 15.365148 15.264489
#> 17 27.410690 27.440299 27.467912 27.493665 27.517682 27.540080 27.560968
#> 18 28.701460 28.586109 28.478532 28.378206 28.284641 28.197382 28.116004
#> 19 21.568258 21.398842 21.240845 21.093496 20.956078 20.827921 20.708402
#> 20 17.024413 16.819995 16.629355 16.451563 16.285754 16.131119 15.986907
#> 21 22.902055 22.937658 22.970862 23.001828 23.030706 23.057639 23.082756
#>       0.6789    0.6331   0.5904   0.5507   0.5135   0.4789   0.4467   0.4166
#> 1  22.139767 22.246514 22.41819 22.57834 22.59614 22.44007 22.44011 22.44014
#> 2  21.326603 21.702978 22.00957 22.29559 22.32736 22.12724 22.12738 22.12749
#> 3  25.415306 25.827209 26.07587 26.30784 26.33361 27.43779 27.43821 27.43851
#> 4  21.623835 21.322756 21.02986 20.75662 20.72626 21.31481 21.31478 21.31475
#> 5  18.500534 17.791665 17.37327 16.98294 16.93960 17.94351 17.94342 17.94335
#> 6  16.876118 15.979990 15.50739 15.06650 15.01755 15.01699 15.01662 15.01635
#> 7  24.805424 25.233897 25.20894 25.18567 25.18306 23.41316 23.41289 23.41270
#> 8  20.052603 19.760296 19.49966 19.25652 19.22951 17.80174 17.80127 17.80095
#> 9  17.528815 17.113984 16.84425 16.59261 16.56466 16.73460 16.73446 16.73435
#> 10  9.237236  9.838036 10.17084 10.48131 10.51579 11.10132 11.10170 11.10197
#> 11  9.302158  9.769237 10.04124 10.29498 10.32316 11.02049 11.02084 11.02109
#> 12 26.725522 27.259488 27.53894 27.79964 27.82859 28.46665 28.46701 28.46727
#> 13 28.867339 28.856907 28.85494 28.85310 28.85289 28.63844 28.63839 28.63835
#> 14 28.906207 29.327119 29.51681 29.69377 29.71342 29.88085 29.88106 29.88120
#> 15 25.845827 25.242969 24.73931 24.26945 24.21726 23.97464 23.97429 23.97404
#> 16 15.170613 14.333315 13.91773 13.53002 13.48698 13.88305 13.88279 13.88261
#> 17 27.580449 27.836863 27.97711 28.10795 28.12248 28.80113 28.80139 28.80157
#> 18 28.040111 27.612508 27.43942 27.27795 27.26002 27.66810 27.66803 27.66797
#> 19 20.596938 20.576587 20.71942 20.85267 20.86748 19.62719 19.62695 19.62678
#> 20 15.852413 16.089261 16.43241 16.75251 16.78808 14.94385 14.94360 14.94344
#> 21 23.106180 23.778422 24.18485 24.56400 24.60611 25.26439 25.26483 25.26514
#>      0.3885   0.3623   0.3379   0.3151   0.2939   0.2741   0.2556   0.2384
#> 1  22.44016 22.44017 22.44019 22.44019 22.44020 22.44020 22.44021 22.44021
#> 2  22.12756 22.12761 22.12765 22.12767 22.12769 22.12770 22.12771 22.12772
#> 3  27.43871 27.43886 27.43896 27.43903 27.43908 27.43912 27.43914 27.43916
#> 4  21.31473 21.31472 21.31470 21.31470 21.31469 21.31469 21.31468 21.31468
#> 5  17.94330 17.94327 17.94324 17.94322 17.94321 17.94320 17.94319 17.94319
#> 6  15.01616 15.01603 15.01593 15.01587 15.01582 15.01579 15.01576 15.01575
#> 7  23.41257 23.41248 23.41241 23.41237 23.41234 23.41232 23.41231 23.41230
#> 8  17.80072 17.80056 17.80045 17.80037 17.80032 17.80028 17.80026 17.80024
#> 9  16.73427 16.73422 16.73418 16.73415 16.73413 16.73412 16.73411 16.73410
#> 10 11.10216 11.10230 11.10239 11.10246 11.10250 11.10254 11.10256 11.10257
#> 11 11.02127 11.02139 11.02148 11.02154 11.02158 11.02161 11.02163 11.02164
#> 12 28.46745 28.46758 28.46766 28.46773 28.46777 28.46780 28.46782 28.46784
#> 13 28.63833 28.63831 28.63830 28.63829 28.63828 28.63828 28.63828 28.63828
#> 14 29.88130 29.88138 29.88143 29.88146 29.88149 29.88150 29.88152 29.88153
#> 15 23.97386 23.97373 23.97364 23.97358 23.97354 23.97351 23.97349 23.97347
#> 16 13.88248 13.88238 13.88232 13.88227 13.88224 13.88221 13.88220 13.88219
#> 17 28.80169 28.80178 28.80184 28.80189 28.80192 28.80194 28.80195 28.80196
#> 18 27.66793 27.66790 27.66788 27.66787 27.66786 27.66785 27.66784 27.66784
#> 19 19.62667 19.62659 19.62653 19.62650 19.62647 19.62645 19.62644 19.62643
#> 20 14.94332 14.94324 14.94319 14.94315 14.94312 14.94311 14.94309 14.94309
#> 21 25.26536 25.26551 25.26562 25.26570 25.26575 25.26579 25.26581 25.26583
#>      0.2223   0.2073   0.1933   0.1803   0.1682   0.1568   0.1463   0.1364
#> 1  22.44021 22.44021 21.69405 21.69385 21.69367 21.69350 21.69334 21.69319
#> 2  22.12772 22.12773 21.67508 21.67491 21.67474 21.67459 21.67445 21.67431
#> 3  27.43917 27.43918 26.94667 26.94673 26.94678 26.94683 26.94688 26.94692
#> 4  21.31468 21.31468 20.94169 20.94167 20.94166 20.94165 20.94163 20.94162
#> 5  17.94318 17.94318 20.31953 20.31989 20.32022 20.32054 20.32083 20.32110
#> 6  15.01574 15.01573 13.96847 13.96830 13.96815 13.96800 13.96786 13.96774
#> 7  23.41229 23.41228 24.23086 24.23088 24.23089 24.23090 24.23091 24.23092
#> 8  17.80022 17.80021 18.92981 18.92994 18.93006 18.93017 18.93028 18.93037
#> 9  16.73410 16.73410 17.65406 17.65410 17.65414 17.65418 17.65422 17.65425
#> 10 11.10259 11.10259 10.65801 10.65801 10.65801 10.65801 10.65802 10.65802
#> 11 11.02165 11.02166 11.02856 11.02865 11.02874 11.02882 11.02890 11.02897
#> 12 28.46785 28.46785 28.98824 28.98832 28.98840 28.98848 28.98855 28.98861
#> 13 28.63827 28.63827 29.47776 29.47780 29.47783 29.47786 29.47788 29.47791
#> 14 29.88153 29.88154 30.92930 30.92936 30.92941 30.92945 30.92950 30.92954
#> 15 23.97346 23.97345 20.32582 20.32536 20.32493 20.32454 20.32417 20.32383
#> 16 13.88218 13.88217 14.14174 14.14184 14.14193 14.14202 14.14210 14.14217
#> 17 28.80197 28.80198 29.02273 29.02279 29.02285 29.02290 29.02295 29.02300
#> 18 27.66784 27.66784 28.48691 28.48725 28.48757 28.48786 28.48814 28.48839
#> 19 19.62642 19.62642 18.58324 18.58307 18.58292 18.58277 18.58264 18.58252
#> 20 14.94308 14.94308 15.46633 15.46621 15.46610 15.46599 15.46590 15.46581
#> 21 25.26584 25.26585 24.03115 24.03107 24.03100 24.03094 24.03088 24.03083
#>      0.1272   0.1186   0.1106   0.1032   0.0962   0.0897   0.0837   0.0781
#> 1  21.69305 21.69292 21.69280 21.69269 21.69259 21.69249 21.69240 21.69232
#> 2  21.67419 21.67408 21.67397 21.67387 21.67378 21.67369 21.67361 21.67354
#> 3  26.94696 26.94699 26.94703 26.94706 26.94709 26.94712 26.94714 26.94717
#> 4  20.94161 20.94160 20.94159 20.94158 20.94157 20.94156 20.94156 20.94155
#> 5  20.32135 20.32158 20.32179 20.32200 20.32218 20.32235 20.32252 20.32266
#> 6  13.96762 13.96751 13.96741 13.96732 13.96723 13.96715 13.96708 13.96701
#> 7  24.23093 24.23094 24.23095 24.23096 24.23096 24.23097 24.23098 24.23098
#> 8  18.93047 18.93055 18.93063 18.93070 18.93077 18.93083 18.93089 18.93095
#> 9  17.65428 17.65431 17.65433 17.65436 17.65438 17.65440 17.65442 17.65444
#> 10 10.65802 10.65802 10.65802 10.65802 10.65803 10.65803 10.65803 10.65803
#> 11 11.02903 11.02909 11.02915 11.02920 11.02925 11.02929 11.02934 11.02938
#> 12 28.98867 28.98872 28.98877 28.98882 28.98887 28.98891 28.98895 28.98898
#> 13 29.47793 29.47795 29.47797 29.47799 29.47801 29.47802 29.47804 29.47805
#> 14 30.92958 30.92961 30.92964 30.92967 30.92970 30.92973 30.92975 30.92977
#> 15 20.32351 20.32321 20.32294 20.32269 20.32245 20.32223 20.32203 20.32184
#> 16 14.14224 14.14230 14.14236 14.14242 14.14247 14.14252 14.14256 14.14260
#> 17 29.02304 29.02308 29.02312 29.02315 29.02318 29.02321 29.02324 29.02327
#> 18 28.48863 28.48885 28.48905 28.48924 28.48942 28.48958 28.48973 28.48987
#> 19 18.58241 18.58230 18.58220 18.58211 18.58203 18.58195 18.58187 18.58181
#> 20 15.46573 15.46565 15.46558 15.46551 15.46545 15.46539 15.46534 15.46529
#> 21 24.03078 24.03073 24.03069 24.03064 24.03061 24.03057 24.03054 24.03051
#>      0.0728   0.0679   0.0633   0.0590   0.0551   0.0514   0.0479   0.0447
#> 1  21.69224 21.69217 21.69210 21.69204 21.69198 21.92036 21.92051 21.92064
#> 2  21.67347 21.67341 21.67335 21.67329 21.67324 21.70836 21.70843 21.70850
#> 3  26.94719 26.94721 26.94723 26.94724 26.94726 26.65237 26.65230 26.65223
#> 4  20.94154 20.94154 20.94153 20.94152 20.94152 21.27258 21.27271 21.27283
#> 5  20.32280 20.32293 20.32305 20.32316 20.32327 20.63292 20.63275 20.63258
#> 6  13.96694 13.96688 13.96682 13.96677 13.96672 14.24474 14.24498 14.24521
#> 7  24.23099 24.23099 24.23100 24.23100 24.23101 24.45537 24.45541 24.45546
#> 8  18.93100 18.93104 18.93109 18.93113 18.93117 18.48892 18.48862 18.48833
#> 9  17.65445 17.65447 17.65448 17.65450 17.65451 16.84076 16.84049 16.84022
#> 10 10.65803 10.65803 10.65803 10.65803 10.65803 11.15154 11.15168 11.15181
#> 11 11.02941 11.02945 11.02948 11.02951 11.02953 11.05025 11.05020 11.05015
#> 12 28.98901 28.98904 28.98907 28.98910 28.98912 28.67690 28.67676 28.67662
#> 13 29.47806 29.47808 29.47809 29.47810 29.47811 30.02555 30.02561 30.02567
#> 14 30.92979 30.92981 30.92983 30.92984 30.92986 30.91966 30.91961 30.91957
#> 15 20.32166 20.32150 20.32135 20.32121 20.32108 20.24951 20.24987 20.25021
#> 16 14.14264 14.14268 14.14271 14.14274 14.14277 13.81475 13.81460 13.81445
#> 17 29.02329 29.02331 29.02333 29.02335 29.02337 28.91833 28.91826 28.91820
#> 18 28.49001 28.49013 28.49024 28.49035 28.49045 28.77184 28.77173 28.77163
#> 19 18.58174 18.58168 18.58163 18.58158 18.58153 18.53259 18.53267 18.53274
#> 20 15.46524 15.46520 15.46516 15.46512 15.46508 15.45618 15.45631 15.45643
#> 21 24.03048 24.03046 24.03043 24.03041 24.03039 23.71653 23.71653 23.71652
#>      0.0417   0.0388   0.0362   0.0338   0.0315   0.0294   0.0274   0.0256
#> 1  21.92078 21.92091 21.92103 21.92116 21.92127 21.92139 21.92150 21.92160
#> 2  21.70856 21.70863 21.70869 21.70875 21.70880 21.70886 21.70891 21.70897
#> 3  26.65216 26.65209 26.65203 26.65197 26.65191 26.65185 26.65180 26.65175
#> 4  21.27294 21.27306 21.27316 21.27327 21.27337 21.27347 21.27356 21.27365
#> 5  20.63243 20.63227 20.63213 20.63199 20.63185 20.63172 20.63159 20.63146
#> 6  14.24543 14.24564 14.24585 14.24605 14.24625 14.24643 14.24661 14.24679
#> 7  24.45550 24.45553 24.45557 24.45561 24.45564 24.45568 24.45571 24.45574
#> 8  18.48805 18.48778 18.48752 18.48727 18.48702 18.48679 18.48656 18.48634
#> 9  16.83997 16.83972 16.83948 16.83925 16.83903 16.83882 16.83861 16.83841
#> 10 11.15194 11.15206 11.15218 11.15230 11.15241 11.15252 11.15262 11.15272
#> 11 11.05010 11.05005 11.05000 11.04996 11.04991 11.04987 11.04983 11.04979
#> 12 28.67649 28.67636 28.67624 28.67612 28.67600 28.67589 28.67578 28.67568
#> 13 30.02573 30.02579 30.02585 30.02590 30.02596 30.02601 30.02606 30.02610
#> 14 30.91953 30.91949 30.91946 30.91942 30.91939 30.91935 30.91932 30.91929
#> 15 20.25054 20.25087 20.25118 20.25148 20.25177 20.25204 20.25232 20.25258
#> 16 13.81432 13.81418 13.81405 13.81393 13.81381 13.81369 13.81358 13.81347
#> 17 28.91814 28.91808 28.91803 28.91797 28.91792 28.91787 28.91782 28.91778
#> 18 28.77152 28.77142 28.77133 28.77123 28.77114 28.77106 28.77097 28.77089
#> 19 18.53281 18.53288 18.53295 18.53301 18.53307 18.53313 18.53319 18.53325
#> 20 15.45654 15.45665 15.45676 15.45686 15.45697 15.45706 15.45716 15.45725
#> 21 23.71652 23.71652 23.71652 23.71651 23.71651 23.71651 23.71651 23.71650
#>      0.0238   0.0222   0.0207   0.0193   0.0180   0.0168   0.0157   0.0146
#> 1  21.92170 21.92180 21.92190 21.92199 21.92208 21.92216 21.92224 21.92232
#> 2  21.70902 21.70906 21.70911 21.70916 21.70920 21.70924 21.70928 21.70932
#> 3  26.65170 26.65165 26.65160 26.65155 26.65151 26.65146 26.65142 26.65138
#> 4  21.27374 21.27382 21.27390 21.27398 21.27406 21.27413 21.27420 21.27427
#> 5  20.63134 20.63123 20.63112 20.63101 20.63090 20.63080 20.63071 20.63061
#> 6  14.24696 14.24712 14.24728 14.24743 14.24758 14.24772 14.24785 14.24799
#> 7  24.45577 24.45580 24.45583 24.45586 24.45588 24.45591 24.45593 24.45596
#> 8  18.48613 18.48593 18.48573 18.48554 18.48535 18.48518 18.48500 18.48484
#> 9  16.83822 16.83803 16.83785 16.83768 16.83751 16.83735 16.83719 16.83704
#> 10 11.15282 11.15291 11.15300 11.15309 11.15317 11.15325 11.15333 11.15341
#> 11 11.04975 11.04971 11.04968 11.04964 11.04961 11.04958 11.04955 11.04952
#> 12 28.67558 28.67548 28.67539 28.67530 28.67521 28.67512 28.67504 28.67496
#> 13 30.02615 30.02619 30.02624 30.02628 30.02632 30.02636 30.02639 30.02643
#> 14 30.91926 30.91923 30.91920 30.91918 30.91915 30.91913 30.91910 30.91908
#> 15 20.25283 20.25307 20.25331 20.25354 20.25375 20.25397 20.25417 20.25437
#> 16 13.81337 13.81326 13.81317 13.81307 13.81298 13.81289 13.81281 13.81272
#> 17 28.91773 28.91769 28.91765 28.91760 28.91757 28.91753 28.91749 28.91746
#> 18 28.77081 28.77073 28.77066 28.77059 28.77052 28.77045 28.77039 28.77033
#> 19 18.53330 18.53335 18.53340 18.53345 18.53350 18.53354 18.53359 18.53363
#> 20 15.45734 15.45742 15.45750 15.45758 15.45766 15.45774 15.45781 15.45788
#> 21 23.71650 23.71650 23.71650 23.71650 23.71649 23.71649 23.71649 23.71649
#>      0.0136   0.0127   0.0119   0.0111   0.0103   0.0096   0.0090   0.0084
#> 1  21.92240 21.92248 21.92255 21.92262 21.92268 21.92275 21.92281 21.92287
#> 2  21.70936 21.70940 21.70943 21.70947 21.70950 21.70953 21.70956 21.70959
#> 3  26.65134 26.65131 26.65127 26.65124 26.65120 26.65117 26.65114 26.65111
#> 4  21.27434 21.27440 21.27446 21.27452 21.27458 21.27463 21.27469 21.27474
#> 5  20.63052 20.63043 20.63035 20.63026 20.63019 20.63011 20.63004 20.62996
#> 6  14.24811 14.24824 14.24836 14.24847 14.24858 14.24869 14.24879 14.24889
#> 7  24.45598 24.45600 24.45602 24.45604 24.45606 24.45608 24.45610 24.45612
#> 8  18.48468 18.48452 18.48437 18.48423 18.48409 18.48395 18.48382 18.48370
#> 9  16.83690 16.83676 16.83662 16.83649 16.83636 16.83624 16.83612 16.83601
#> 10 11.15348 11.15355 11.15362 11.15368 11.15375 11.15381 11.15387 11.15392
#> 11 11.04949 11.04946 11.04943 11.04941 11.04938 11.04936 11.04933 11.04931
#> 12 28.67489 28.67481 28.67474 28.67468 28.67461 28.67455 28.67448 28.67442
#> 13 30.02646 30.02650 30.02653 30.02656 30.02659 30.02662 30.02665 30.02667
#> 14 30.91906 30.91903 30.91901 30.91899 30.91897 30.91895 30.91894 30.91892
#> 15 20.25456 20.25475 20.25492 20.25510 20.25526 20.25542 20.25558 20.25573
#> 16 13.81265 13.81257 13.81250 13.81242 13.81235 13.81229 13.81222 13.81216
#> 17 28.91742 28.91739 28.91736 28.91733 28.91730 28.91727 28.91724 28.91721
#> 18 28.77027 28.77021 28.77015 28.77010 28.77005 28.77000 28.76995 28.76990
#> 19 18.53367 18.53371 18.53375 18.53378 18.53382 18.53385 18.53388 18.53392
#> 20 15.45794 15.45801 15.45807 15.45813 15.45819 15.45825 15.45830 15.45835
#> 21 23.71649 23.71649 23.71649 23.71649 23.71648 23.71648 23.71648 23.71648
#>      0.0078   0.0073   0.0068   0.0063   0.0059   0.0055
#> 1  21.92293 21.92299 21.92304 21.92309 21.92314 21.92319
#> 2  21.70962 21.70965 21.70968 21.70970 21.70973 21.70975
#> 3  26.65108 26.65105 26.65102 26.65100 26.65097 26.65095
#> 4  21.27479 21.27484 21.27488 21.27493 21.27497 21.27501
#> 5  20.62989 20.62983 20.62976 20.62970 20.62964 20.62958
#> 6  14.24899 14.24908 14.24917 14.24926 14.24935 14.24943
#> 7  24.45614 24.45615 24.45617 24.45618 24.45620 24.45621
#> 8  18.48357 18.48346 18.48334 18.48323 18.48313 18.48303
#> 9  16.83590 16.83580 16.83569 16.83559 16.83550 16.83541
#> 10 11.15398 11.15403 11.15408 11.15413 11.15418 11.15423
#> 11 11.04929 11.04927 11.04925 11.04923 11.04921 11.04919
#> 12 28.67437 28.67431 28.67426 28.67421 28.67416 28.67411
#> 13 30.02670 30.02672 30.02675 30.02677 30.02679 30.02681
#> 14 30.91890 30.91889 30.91887 30.91885 30.91884 30.91883
#> 15 20.25587 20.25601 20.25615 20.25628 20.25640 20.25653
#> 16 13.81210 13.81204 13.81199 13.81193 13.81188 13.81183
#> 17 28.91719 28.91716 28.91714 28.91711 28.91709 28.91707
#> 18 28.76985 28.76981 28.76977 28.76973 28.76969 28.76965
#> 19 18.53395 18.53398 18.53401 18.53403 18.53406 18.53409
#> 20 15.45840 15.45845 15.45850 15.45855 15.45859 15.45863
#> 21 23.71648 23.71648 23.71648 23.71648 23.71648 23.71648
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 22.8 21.4 18.7 14.3 22.8 17.8 17.3 10.4 14.7 32.4 30.4 33.9 21.5
#> [16] 13.3 27.3 30.4 19.7 15.0 21.4
#> 
#> $X
#>               am        carb        cyl       disp        drat       gear
#>  [1,]  0.9534626  0.60070445  0.1097643 -0.3719818  0.39982591  0.3483665
#>  [2,]  0.9534626  0.60070445  0.1097643 -0.3719818  0.39982591  0.3483665
#>  [3,]  0.9534626 -1.04470340 -1.0427605 -0.7920915  0.29190375  0.3483665
#>  [4,] -1.0488088 -1.04470340  0.1097643  0.4197635 -1.37009754 -1.1147728
#>  [5,] -1.0488088 -0.49623411  1.2622890  1.2438249 -1.21900652 -1.1147728
#>  [6,] -1.0488088  0.60070445  1.2622890  1.2438249 -1.08949992 -1.1147728
#>  [7,] -1.0488088 -0.49623411 -1.0427605 -0.5270992  0.44299478  0.3483665
#>  [8,] -1.0488088  0.60070445  0.1097643 -0.3105811  0.44299478  0.3483665
#>  [9,] -1.0488088  0.05223517  1.2622890  0.5635703 -1.39168197 -1.1147728
#> [10,] -1.0488088  0.60070445  1.2622890  2.0517282 -1.54277300 -1.1147728
#> [11,] -1.0488088  0.60070445  1.2622890  1.8901476 -1.04633106 -1.1147728
#> [12,]  0.9534626 -1.04470340 -1.0427605 -1.0288072  0.78834570  0.3483665
#> [13,]  0.9534626 -0.49623411 -1.0427605 -1.0530443  2.62302245  0.3483665
#> [14,]  0.9534626 -1.04470340 -1.0427605 -1.0902079  1.09052775  0.3483665
#> [15,] -1.0488088 -1.04470340 -1.0427605 -0.6943352 -0.03186273 -1.1147728
#> [16,] -1.0488088  0.60070445  1.2622890  1.1630346  0.03289056 -1.1147728
#> [17,]  0.9534626 -1.04470340 -1.0427605 -1.0263835  0.78834570  0.3483665
#> [18,]  0.9534626 -0.49623411 -1.0427605 -0.8963111  0.11922829  1.8115058
#> [19,]  0.9534626  1.69764302  0.1097643 -0.4931673 -0.20453819  1.8115058
#> [20,]  0.9534626  2.79458158  1.2622890  0.7671619 -0.37721365  1.8115058
#> [21,]  0.9534626 -0.49623411 -1.0427605 -0.6870641  0.85309899  0.3483665
#>               hp        qsec         vs         wt
#>  [1,] -0.4543531 -0.79273357 -1.0488088 -0.4018750
#>  [2,] -0.4543531 -0.49374233 -1.0488088 -0.1525654
#>  [3,] -0.6864046  0.35517922  0.9534626 -0.6951803
#>  [4,] -0.4543531  0.79832694  0.9534626  0.1798474
#>  [5,]  0.4329029 -0.49374233 -1.0488088  0.3998265
#>  [6,]  1.3884094 -1.12375958 -1.0488088  0.5269255
#>  [7,] -0.6591044  2.64566565  0.9534626  0.1162979
#>  [8,] -0.2769019  0.51001396  0.9534626  0.3998265
#>  [9,]  0.5011534 -0.18407284 -1.0488088  0.6833550
#> [10,]  0.9789066 -0.06661199 -1.0488088  2.3395528
#> [11,]  1.1836580 -0.28017716 -1.0488088  2.2623157
#> [12,] -1.0549571  0.81434433  0.9534626 -0.8125025
#> [13,] -1.2460584  0.30712705  0.9534626 -1.3844480
#> [14,] -1.0686072  1.04392689  0.9534626 -1.1693574
#> [15,] -0.6318043  1.10265731  0.9534626 -0.5534161
#> [16,]  1.3884094 -1.35334214 -1.0488088  0.7909003
#> [17,] -1.0549571  0.51001396  0.9534626 -1.0715889
#> [18,] -0.4134028 -0.55781188  0.9534626 -1.4841719
#> [19,]  0.4329029 -1.30528997 -1.0488088 -0.2552223
#> [20,]  2.6169176 -1.78581160 -1.0488088  0.5269255
#> [21,] -0.4680032  0.34984009  0.9534626 -0.2454454
#> attr(,"center")
#>  [1]   0.5238095   2.9047619   5.8095238 206.0428571   3.7147619   3.7619048
#>  [7] 143.2857143  17.9447619   0.5238095   3.0310476
#> attr(,"scale")
#>  [1]   0.4994328   1.8232562   1.7353207 123.7771830   0.4632969   0.6834619
#>  [7]  73.2595795   1.8729646   0.4994328   1.0228247
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 44
#> 
#> $lambda.min
#> [1] 0.2740625
#> 
#> $null.dev
#> [1] 40.04562
#> 
#> $Bias
#> [1] 4.708966
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 5.572117