Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 48.74300 45.35662 40.10768 35.28330 31.35559 28.18967 25.66947 23.69515
#>   [9] 22.18094 21.05324 20.24893 19.71401 19.40229 19.27436 19.29661 19.39731
#>  [17] 18.96531 18.77329 18.76416 18.76416 18.75412 18.67431 18.64478 18.45426
#>  [25] 18.29627 17.49854 19.42350 19.52310 19.46833 18.39002 17.28223 16.04663
#>  [33] 15.96319 15.80817 13.36215 13.60682 13.74720 13.64748 13.87599 13.71897
#>  [41] 13.61886 12.52906 12.73715 12.75017 12.74958 12.74914 12.84376 12.84361
#>  [49] 12.84966 14.31670 16.66721 16.62925 16.31131 15.97509 18.17357 18.13813
#>  [57] 18.50069 18.28543 17.57943 17.35786 16.79121 17.12885 15.46390 15.39882
#>  [65] 14.74363 14.41945 14.53456 14.53456 14.53455 14.53460 14.72766 15.14737
#>  [73] 15.33098 15.33058 15.33017 15.48620 15.48598 15.48576 15.37047 15.37030
#>  [81] 14.97901 14.97889 14.97878 14.97865 14.97822 14.97801 14.97780 14.97759
#>  [89] 14.97738 14.95886 14.95861 15.12542 15.15642 15.15673 15.15702 15.15729
#>  [97] 15.15755 15.15778 15.15800 15.15821
#> 
#> $cvse
#>   [1] 12.518185 12.121221 10.932809  9.730341  8.675143  7.745933  6.927929
#>   [8]  6.212032  5.594050  5.073815  4.653962  4.338082  4.128224  4.022299
#>  [15]  4.012546  4.082404  3.990085  3.915352  3.909337  3.909337  3.911048
#>  [22]  3.928413  3.943898  3.940976  3.933618  3.777541  4.861395  4.801388
#>  [29]  4.904553  5.000144  4.851121  4.633734  4.602236  4.583146  4.547828
#>  [36]  4.512420  4.502187  4.500308  4.500947  4.464369  4.444814  4.333871
#>  [43]  4.335285  4.337394  4.337385  4.337377  4.324664  4.324648  4.326762
#>  [50]  4.373448  4.983284  4.986654  4.840004  5.058687  4.921866  4.937446
#>  [57]  4.886711  4.884413  4.937554  4.955696  4.715620  4.665020  4.718149
#>  [64]  4.727238  4.495124  4.532225  4.522541  4.522429  4.522316  4.522198
#>  [71]  4.516228  4.820068  4.802896  4.802800  4.802711  4.819578  4.819518
#>  [78]  4.819464  4.773927  4.773907  4.780974  4.780949  4.780924  4.780901
#>  [85]  4.780893  4.780883  4.780873  4.780864  4.780855  4.783272  4.783270
#>  [92]  4.780732  4.778379  4.778378  4.778378  4.778378  4.778379  4.778380
#>  [99]  4.778381  4.778382
#> 
#> $fold
#>  [1]  1 10  6  7  9  2  7  3  5  4 10  3  4  8  1  5  2  8  9  6  1
#> 
#> $lambda
#>   [1] 6.140780213 5.726912179 5.340937465 4.980976156 4.645275034 4.332199044
#>   [7] 4.040223327 3.767925797 3.513980209 3.277149704 3.056280782 2.850297687
#>  [13] 2.658197162 2.479043570 2.311964331 2.156145673 2.010828671 1.875305548
#>  [19] 1.748916231 1.631045130 1.521118147 1.418599875 1.322990991 1.233825827
#>  [25] 1.150670095 1.073118782 1.000794168 0.933343990 0.870439729 0.811775005
#>  [31] 0.757064086 0.706040501 0.658455734 0.614078021 0.572691218 0.534093747
#>  [37] 0.498097616 0.464527503 0.433219904 0.404022333 0.376792580 0.351398021
#>  [43] 0.327714970 0.305628078 0.285029769 0.265819716 0.247904357 0.231196433
#>  [49] 0.215614567 0.201082867 0.187530555 0.174891623 0.163104513 0.152111815
#>  [55] 0.141859988 0.132299099 0.123382583 0.115067010 0.107311878 0.100079417
#>  [61] 0.093334399 0.087043973 0.081177500 0.075706409 0.070604050 0.065845573
#>  [67] 0.061407802 0.057269122 0.053409375 0.049809762 0.046452750 0.043321990
#>  [73] 0.040402233 0.037679258 0.035139802 0.032771497 0.030562808 0.028502977
#>  [79] 0.026581972 0.024790436 0.023119643 0.021561457 0.020108287 0.018753055
#>  [85] 0.017489162 0.016310451 0.015211181 0.014185999 0.013229910 0.012338258
#>  [91] 0.011506701 0.010731188 0.010007942 0.009333440 0.008704397 0.008117750
#>  [97] 0.007570641 0.007060405 0.006584557 0.006140780
#> 
#> $fit
#> $beta
#>               6.1408     5.7269    5.3409    4.9810    4.6453    4.3322
#> (Intercept) 20.19048 21.9968269 23.681436 25.252507 26.717694 28.084132
#> am           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000 -0.5538364 -1.070346 -1.552045 -2.001279 -2.420236
#>                4.0402    3.7679    3.5140    3.2771    3.0563   2.8503
#> (Intercept) 29.358477 30.546935 31.655295 32.688955 33.652950 34.55198
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000
#> wt          -2.810956 -3.175344 -3.515172 -3.832098 -4.127664 -4.40331
#>                2.6582   2.4790    2.3120    2.1561    2.0108    1.8753
#> (Intercept) 35.390409 36.17234 36.901562 37.581641 38.058334 38.058334
#> am           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.00000  0.000000  0.000000  0.000000  0.000000
#> wt          -4.660378 -4.90012 -5.123705 -5.332221 -5.478377 -5.478377
#>                1.7489    1.6310    1.5211      1.4186     1.3230     1.2338
#> (Intercept) 38.058334 38.058334 38.058334 37.47101935 35.7915038 34.2253901
#> am           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> carb         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> cyl          0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> disp         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> qsec         0.000000  0.000000  0.000000  0.03066909  0.1183721  0.2001533
#> vs           0.000000  0.000000  0.000000  0.00000000  0.0000000  0.0000000
#> wt          -5.478377 -5.478377 -5.478377 -5.46718681 -5.4351864 -5.4053466
#>                 1.1507     1.0731     1.0008    0.9333     0.8704    0.8118
#> (Intercept) 32.7648272 31.4027014 30.1329656 28.948219 25.9340323 13.354031
#> am           0.0000000  0.0000000  0.0000000  0.000000  0.5201400  3.663872
#> carb         0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.000000
#> cyl          0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.000000
#> disp         0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.000000
#> drat         0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.000000
#> gear         0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.000000
#> hp           0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.000000
#> qsec         0.2764227  0.3475518  0.4138564  0.475723  0.5958335  1.027383
#> vs           0.0000000  0.0000000  0.0000000  0.000000  0.0000000  0.000000
#> wt          -5.3775180 -5.3515649 -5.3273722 -5.304799 -5.1027910 -3.989273
#>                0.7571    0.7060    0.6585    0.6141    0.5727     0.5341
#> (Intercept)  7.668252  7.248670  7.246099  7.244547  7.243609  8.8210325
#> am           4.949945  4.987540  4.988269  4.988708  4.988974  5.2907371
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000 -0.2516268
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> qsec         1.232093  1.251308  1.251390  1.251440  1.251470  1.1498229
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.0000000
#> wt          -3.523464 -3.505020 -3.504769 -3.504617 -3.504525 -3.2469309
#>                 0.4981     0.4645     0.4332     0.4040     0.3768     0.3514
#> (Intercept) 12.0713687 12.8165242 12.8157115 12.8151236 12.8146987 12.8143917
#> am           5.9161740  6.0587008  6.0594400  6.0599739  6.0603594  6.0606378
#> carb        -0.7714104 -0.8902982 -0.8904751 -0.8906028 -0.8906949 -0.8907615
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.9401268  0.8921118  0.8921029  0.8920966  0.8920920  0.8920887
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.7140854 -2.5923892 -2.5920250 -2.5917620 -2.5915721 -2.5914350
#>                 0.3277     0.3056     0.2850     0.2658     0.2479     0.2312
#> (Intercept) 12.8141699 12.8140098 12.8138941 12.8138106 12.8137503 12.8137068
#> am           6.0608388  6.0609840  6.0610888  6.0611644  6.0612191  6.0612585
#> carb        -0.8908095 -0.8908442 -0.8908692 -0.8908873 -0.8909003 -0.8909098
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.8920863  0.8920846  0.8920833  0.8920824  0.8920818  0.8920813
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.5913360 -2.5912645 -2.5912129 -2.5911756 -2.5911487 -2.5911293
#>                 0.2156     0.2011     0.1875     0.1749     0.1631     0.1521
#> (Intercept) 12.8136754  9.5076253  9.5075618  9.5076929 12.3436645 12.3453593
#> am           6.0612870  5.5887524  5.5880188  5.5875068  5.5942355  5.5940135
#> carb        -0.8909166 -1.0058660 -1.0058514 -1.0058267 -1.0688208 -1.0690071
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  1.2989390  1.2993806  1.2996248  1.3641397  1.3647441
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.8920810  0.7747339  0.7746792  0.7746477  0.5073855  0.5071021
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  1.4436890  1.4445471
#> wt          -2.5911153 -2.2161156 -2.2162111 -2.2163100 -1.8218191 -1.8213745
#>                 0.1419     0.1323     0.1234     0.1151      0.1073      0.1001
#> (Intercept) 12.3467651 12.3479308 12.3488972 -13.902774 -13.9075613 -13.9120202
#> am           5.5938286  5.5936749  5.5935470   5.743936   5.7440797   5.7442138
#> carb        -1.0691616 -1.0692898 -1.0693960  -1.827761  -1.8279578  -1.8281415
#> cyl          0.0000000  0.0000000  0.0000000   1.639141   1.6394860   1.6398073
#> disp         0.0000000  0.0000000  0.0000000   0.000000   0.0000000   0.0000000
#> drat         1.3652456  1.3656618  1.3660071   2.819151   2.8194077   2.8196468
#> gear         0.0000000  0.0000000  0.0000000   1.881253   1.8816188   1.8819592
#> hp           0.0000000  0.0000000  0.0000000   0.000000   0.0000000   0.0000000
#> qsec         0.5068671  0.5066722  0.5065106   0.787377   0.7874025   0.7874262
#> vs           1.4452585  1.4458483  1.4463372   2.702860   2.7032298   2.7035745
#> wt          -1.8210058 -1.8207003 -1.8204470  -1.656091  -1.6560034  -1.6559221
#>                  0.0933       0.0870       0.0812       0.0757       0.0706
#> (Intercept) -13.9161728 -21.16710023 -21.17119088 -21.17511918 -21.17888794
#> am            5.7443387   6.04346971   6.04366526   6.04386614   6.04406954
#> carb         -1.8283125  -2.20628636  -2.20640616  -2.20652302  -2.20663721
#> cyl           1.6401065   1.69671656   1.69704729   1.69735440   1.69763911
#> disp          0.0000000   0.00000000   0.00000000   0.00000000   0.00000000
#> drat          2.8198695   3.39150329   3.39161440   3.39172475   3.39183492
#> gear          1.8822763   1.98451973   1.98476600   1.98498839   1.98518968
#> hp            0.0000000   0.01414655   0.01414646   0.01414667   0.01414713
#> qsec          0.7874483   0.98113408   0.98118999   0.98124785   0.98130683
#> vs            2.7038956   2.59573502   2.59606433   2.59636069   2.59662694
#> wt           -1.6558463  -1.70433232  -1.70436880  -1.70440608  -1.70444303
#>                   0.0658       0.0614       0.0573       0.0534       0.0498
#> (Intercept) -21.18249998 -21.18595816 -21.18926547 -21.19242501 -21.19544002
#> am            6.04427304   6.04447457   6.04467237   6.04486499   6.04505125
#> carb         -2.20674897  -2.20685842  -2.20696564  -2.20707066  -2.20717347
#> cyl           1.69790268   1.69814632   1.69837125   1.69857864   1.69876965
#> disp          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> drat          3.39194532   3.39205620   3.39216764   3.39227964   3.39239210
#> gear          1.98537235   1.98553861   1.98569045   1.98582964   1.98595772
#> hp            0.01414782   0.01414871   0.01414977   0.01415096   0.01415227
#> qsec          0.98136621   0.98142536   0.98148376   0.98154096   0.98159659
#> vs            2.59686575   2.59707965   2.59727100   2.59744200   2.59759469
#> wt           -1.70447874  -1.70451250  -1.70454373  -1.70457202  -1.70459709
#>                   0.0465       0.0433       0.0404      0.0377       0.0351
#> (Intercept) -21.19831389 -21.20105016 -21.20365251 -21.2061248 -21.20847086
#> am            6.04523021   6.04540113   6.04556350   6.0457169   6.04586123
#> carb         -2.20727405  -2.20737234  -2.20746828  -2.2075618  -2.20765282
#> cyl           1.69894540   1.69910696   1.69925534   1.6993915   1.69951649
#> disp          0.00000000   0.00000000   0.00000000   0.0000000   0.00000000
#> drat          3.39250484   3.39261764   3.39273025   3.3928424   3.39295375
#> gear          1.98607607   1.98618591   1.98628827   1.9863841   1.98647415
#> hp            0.01415367   0.01415514   0.01415665   0.0141582   0.01415976
#> qsec          0.98165034   0.98170198   0.98175132   0.9817982   0.98184262
#> vs            2.59773095   2.59785253   2.59796103   2.5980579   2.59814445
#> wt           -1.70461875  -1.70463692  -1.70465158  -1.7046628  -1.70467064
#>                   0.0328       0.0306       0.0285        0.0266       0.0248
#> (Intercept) -21.21069483 -21.21280082 -21.21479301 -2.122510e+01 -21.54653953
#> am            6.04599629   6.04612216   6.04623894  6.047190e+00   6.04741033
#> carb         -2.20774128  -2.20782710  -2.20791023 -2.208229e+00  -2.08587670
#> cyl           1.69963107   1.69973610   1.69983236  1.699946e+00   1.64113587
#> disp          0.00000000   0.00000000   0.00000000  9.202733e-06   0.00461143
#> drat          3.39306406   3.39317301   3.39328033  3.394565e+00   3.35628580
#> gear          1.98655913   1.98663963   1.98671614  1.986934e+00   1.98654813
#> hp            0.01416133   0.01416288   0.01416441  1.416517e-02   0.01137428
#> qsec          0.98188443   0.98192366   0.98196031  9.820895e-01   1.04214526
#> vs            2.59822191   2.59829135   2.59835375  2.598528e+00   2.58223925
#> wt           -1.70467528  -1.70467690  -1.70467570 -1.704673e+00  -2.09183524
#>                    0.0231        0.0216        0.0201        0.0188
#> (Intercept) -21.540638591 -21.535033581 -21.529714624 -21.524671872
#> am            6.047288072   6.047167090   6.047047666   6.046930056
#> carb         -2.085657127  -2.085449652  -2.085253758  -2.085068936
#> cyl           1.640797503   1.640479526   1.640181043   1.639901185
#> disp          0.004610458   0.004609456   0.004608431   0.004607388
#> drat          3.355925703   3.355586414   3.355267018   3.354966622
#> gear          1.986149964   1.985775277   1.985423047   1.985092279
#> hp            0.011373067   0.011371874   0.011370698   0.011369540
#> qsec          1.042086929   1.042029115   1.041971969   1.041915627
#> vs            2.581878842   2.581541906   2.581227332   2.580934039
#> wt           -2.091840966  -2.091837901  -2.091826984  -2.091809096
#>                    0.0175        0.0163        0.0152        0.0142
#> (Intercept) -21.519895528 -21.515375864 -21.511103251 -21.507068167
#> am            6.046814486   6.046701158   6.046590248   6.046481907
#> carb         -2.084894694  -2.084730553  -2.084576048  -2.084430729
#> cyl           1.639639101   1.639393962   1.639164961   1.638951314
#> disp          0.004606335   0.004605275   0.004604215   0.004603159
#> drat          3.354684358   3.354419384   3.354170882   3.353938058
#> gear          1.984782001   1.984491266   1.984219155   1.983964772
#> hp            0.011368401   0.011367282   0.011366184   0.011365107
#> qsec          1.041860213   1.041805836   1.041752590   1.041700560
#> vs            2.580660975   2.580407120   2.580171486   2.579953115
#> wt           -2.091785057  -2.091755628  -2.091721514  -2.091683368
#>                   0.0132        0.0123       0.0115        0.0107        0.0100
#> (Intercept) -21.50326123 -21.499673186 -21.49629497 -21.493117661 -21.490132546
#> am            6.04637627   6.046273440   6.04617351   6.046076567   6.045982655
#> carb         -2.08429416  -2.084165923  -2.08404561  -2.083932821  -2.083827184
#> cyl           1.63875226   1.638567058   1.63839500   1.638235385   1.638087554
#> disp          0.00460211   0.004601073   0.00460005   0.004599044   0.004598057
#> drat          3.35372014   3.353516400   3.35332611   3.353148568   3.352983120
#> gear          1.98372725   1.983505744   1.98329944   1.983107559   1.982929330
#> hp            0.01136405   0.011363022   0.01136201   0.011361029   0.011360070
#> qsec          1.04164982   1.041600418   1.04155242   1.041505855   1.041460764
#> vs            2.57975108   2.579564497   2.57939250   2.579234253   2.579088970
#> wt           -2.09164179  -2.091597341  -2.09155053  -2.091501814  -2.091451633
#>                    0.0093        0.0087        0.0081        0.0076
#> (Intercept) -21.487331094 -21.484704978 -21.482246084 -21.479946511
#> am            6.045891819   6.045804090   6.045719482   6.045637999
#> carb         -2.083728333  -2.083635914  -2.083549591  -2.083469036
#> cyl           1.637950861   1.637824685   1.637708430   1.637601521
#> disp          0.004597093   0.004596151   0.004595236   0.004594346
#> drat          3.352829119   3.352685943   3.352553000   3.352429716
#> gear          1.982764023   1.982610930   1.982469372   1.982338694
#> hp            0.011359134   0.011358224   0.011357338   0.011356478
#> qsec          1.041417167   1.041375082   1.041334519   1.041295481
#> vs            2.578955881   2.578834252   2.578723379   2.578622589
#> wt           -2.091400374  -2.091348390  -2.091296004  -2.091243504
#>                    0.0071        0.0066        0.0061
#> (Intercept) -21.477798581 -21.475794837 -21.473928054
#> am            6.045559635   6.045484372   6.045412185
#> carb         -2.083393939  -2.083323999  -2.083258929
#> cyl           1.637503409   1.637413564   1.637331482
#> disp          0.004593485   0.004592651   0.004591848
#> drat          3.352315543   3.352209956   3.352112450
#> gear          1.982218268   1.982107491   1.982005786
#> hp            0.011355643   0.011354833   0.011354048
#> qsec          1.041257968   1.041221971   1.041187480
#> vs            2.578531237   2.578448708   2.578374415
#> wt           -2.091191153  -2.091139182  -2.091087800
#> 
#> $iter
#>   [1]   0   8   4   3   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   4   4   4   4   4   3   3  48  69  17   7   1   1   1  65
#>  [37]  77  15   1   1   1   1   1   1   1   1   1   1   1  65   1   1  36   1
#>  [55]   1   1   1 129   1   1   1  78   1   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   2 100   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 6.140780213 5.726912179 5.340937465 4.980976156 4.645275034 4.332199044
#>   [7] 4.040223327 3.767925797 3.513980209 3.277149704 3.056280782 2.850297687
#>  [13] 2.658197162 2.479043570 2.311964331 2.156145673 2.010828671 1.875305548
#>  [19] 1.748916231 1.631045130 1.521118147 1.418599875 1.322990991 1.233825827
#>  [25] 1.150670095 1.073118782 1.000794168 0.933343990 0.870439729 0.811775005
#>  [31] 0.757064086 0.706040501 0.658455734 0.614078021 0.572691218 0.534093747
#>  [37] 0.498097616 0.464527503 0.433219904 0.404022333 0.376792580 0.351398021
#>  [43] 0.327714970 0.305628078 0.285029769 0.265819716 0.247904357 0.231196433
#>  [49] 0.215614567 0.201082867 0.187530555 0.174891623 0.163104513 0.152111815
#>  [55] 0.141859988 0.132299099 0.123382583 0.115067010 0.107311878 0.100079417
#>  [61] 0.093334399 0.087043973 0.081177500 0.075706409 0.070604050 0.065845573
#>  [67] 0.061407802 0.057269122 0.053409375 0.049809762 0.046452750 0.043321990
#>  [73] 0.040402233 0.037679258 0.035139802 0.032771497 0.030562808 0.028502977
#>  [79] 0.026581972 0.024790436 0.023119643 0.021561457 0.020108287 0.018753055
#>  [85] 0.017489162 0.016310451 0.015211181 0.014185999 0.013229910 0.012338258
#>  [91] 0.011506701 0.010731188 0.010007942 0.009333440 0.008704397 0.008117750
#>  [97] 0.007570641 0.007060405 0.006584557 0.006140780
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 28
#> 
#> $loss
#>   [1] 999.77810 847.75862 720.57172 614.64380 526.88959 454.64697 395.62046
#>   [8] 347.83212 309.57890 279.39573 256.02337 238.38066 225.54032 216.70804
#>  [15] 211.20435 208.44892 207.88528 207.88528 207.88528 207.88528 207.88528
#>  [22] 204.62126 195.93901 188.71307 182.73085 177.80996 173.79488 170.54631
#>  [29] 156.58637 112.73095 108.08810 108.06801 108.06800 108.06800 108.06800
#>  [36]  99.79676  91.32406  91.01720  91.01719  91.01719  91.01719  91.01719
#>  [43]  91.01718  91.01718  91.01718  91.01718  91.01718  91.01718  91.01718
#>  [50]  88.11017  88.11017  88.11017  85.49945  85.49943  85.49942  85.49941
#>  [57]  85.49941  75.66415  75.66414  75.66413  75.66412  73.76893  73.76892
#>  [64]  73.76891  73.76890  73.76890  73.76889  73.76889  73.76888  73.76888
#>  [71]  73.76887  73.76887  73.76887  73.76886  73.76886  73.76886  73.76886
#>  [78]  73.76886  73.76749  73.42876  73.42874  73.42873  73.42872  73.42870
#>  [85]  73.42870  73.42869  73.42868  73.42867  73.42867  73.42866  73.42866
#>  [92]  73.42865  73.42865  73.42865  73.42864  73.42864  73.42864  73.42864
#>  [99]  73.42864  73.42864
#> 
#> $linear.predictors
#>      6.1408   5.7269   5.3409   4.9810   4.6453   4.3322   4.0402   3.7679
#> 1  20.19048 20.40455 20.60419 20.79038 20.96402 21.12595 21.27698 21.41782
#> 2  20.19048 20.71193 21.19823 21.65176 22.07473 22.46919 22.83706 23.18014
#> 3  20.19048 20.21624 20.24027 20.26268 20.28358 20.30307 20.32125 20.33821
#> 4  20.19048 20.09163 19.99944 19.91347 19.83330 19.75852 19.68879 19.62375
#> 5  20.19048 20.01963 19.86030 19.71171 19.57313 19.44389 19.32336 19.21096
#> 6  20.19048 20.25224 20.30985 20.36357 20.41367 20.46039 20.50396 20.54460
#> 7  20.19048 20.09163 19.99944 19.91347 19.83330 19.75852 19.68879 19.62375
#> 8  20.19048 19.74271 19.32513 18.93568 18.57249 18.23377 17.91789 17.62329
#> 9  20.19048 19.90333 19.63553 19.38578 19.15286 18.93564 18.73306 18.54414
#> 10 20.19048 19.08919 18.06212 17.10427 16.21098 15.37790 14.60096 13.87638
#> 11 20.19048 18.99282 17.87588 16.83422 15.86276 14.95677 14.11185 13.32387
#> 12 20.19048 19.03657 17.96044 16.95683 16.02086 15.14797 14.33392 13.57472
#> 13 20.19048 20.77839 21.32667 21.83801 22.31488 22.75961 23.17437 23.56118
#> 14 20.19048 21.10238 21.95283 22.74595 23.48563 24.17545 24.81878 25.41876
#> 15 20.19048 20.98054 21.71735 22.40451 23.04535 23.64300 24.20037 24.72018
#> 16 20.19048 20.63162 21.04303 21.42672 21.78454 22.11825 22.42947 22.71971
#> 17 20.19048 20.09440 20.00480 19.92123 19.84330 19.77062 19.70284 19.63963
#> 18 20.19048 19.87010 19.57131 19.29266 19.03278 18.79043 18.56440 18.35362
#> 19 20.19048 20.81162 21.39089 21.93113 22.43496 22.90483 23.34303 23.75170
#> 20 20.19048 21.15887 22.06200 22.90426 23.68976 24.42232 25.10550 25.74264
#> 21 20.19048 20.01963 19.86030 19.71171 19.57313 19.44389 19.32336 19.21096
#>      3.5140   3.2771   3.0563   2.8503   2.6582    2.4790    2.3120    2.1561
#> 1  21.54917 21.67167 21.78592 21.89246 21.99182 22.084489 22.170910 22.251506
#> 2  23.50009 23.79849 24.07677 24.33630 24.57833 24.804056 25.014566 25.210889
#> 3  20.35402 20.36876 20.38251 20.39533 20.40729 20.418448 20.428850 20.438551
#> 4  19.56310 19.50654 19.45379 19.40459 19.35871 19.315921 19.276017 19.238802
#> 5  19.10613 19.00837 18.91719 18.83216 18.75286 18.678906 18.609935 18.545613
#> 6  20.58250 20.61785 20.65081 20.68155 20.71022 20.736956 20.761891 20.785146
#> 7  19.56310 19.50654 19.45379 19.40459 19.35871 19.315921 19.276017 19.238802
#> 8  17.34854 17.09232 16.85336 16.63051 16.42267 16.228846 16.048083 15.879503
#> 9  18.36794 18.20362 18.05038 17.90747 17.77418 17.649881 17.533957 17.425847
#> 10 13.20064 12.57044 11.98271 11.43460 10.92343 10.446704 10.002111  9.587482
#> 11 12.58900 11.90366 11.26450 10.66842 10.11252  9.594083  9.110586  8.659676
#> 12 12.86670 12.20639 11.59059 11.01629 10.48069  9.981192  9.515359  9.080921
#> 13 23.92192 24.25834 24.57209 24.86469 25.13758 25.392070 25.629411 25.850756
#> 14 25.97829 26.50012 26.98677 27.44063 27.86390 28.258641 28.626778 28.970105
#> 15 25.20495 25.65706 26.07869 26.47190 26.83862 27.180614 27.499563 27.797016
#> 16 22.99039 23.24283 23.47826 23.69782 23.90258 24.093539 24.271629 24.437717
#> 17 19.58068 19.52570 19.47442 19.42661 19.38201 19.340422 19.301635 19.265463
#> 18 18.15703 17.97370 17.80272 17.64327 17.49456 17.355873 17.226535 17.105913
#> 19 24.13283 24.48827 24.81975 25.12889 25.41720 25.686078 25.936833 26.170689
#> 20 26.33684 26.89099 27.40779 27.88977 28.33926 28.758453 29.149396 29.513991
#> 21 19.10613 19.00837 18.91719 18.83216 18.75286 18.678906 18.609935 18.545613
#>       2.0108    1.8753    1.7489    1.6310    1.5211    1.4186    1.3230
#> 1  22.307999 22.307999 22.307999 22.307999 22.307999 22.274845 22.180035
#> 2  25.348499 25.348499 25.348499 25.348499 25.348499 25.357898 25.384775
#> 3  20.445351 20.445351 20.445351 20.445351 20.445351 20.490221 20.618532
#> 4  19.212716 19.212716 19.212716 19.212716 19.212716 19.185885 19.109155
#> 5  18.500527 18.500527 18.500527 18.500527 18.500527 18.438961 18.262902
#> 6  20.801446 20.801446 20.801446 20.801446 20.801446 20.951703 21.381387
#> 7  19.212716 19.212716 19.212716 19.212716 19.212716 19.243542 19.331695
#> 8  15.761339 15.761339 15.761339 15.761339 15.761339 15.753211 15.729969
#> 9  17.350068 17.350068 17.350068 17.350068 17.350068 17.357097 17.377196
#> 10  9.296854  9.296854  9.296854  9.296854  9.296854  9.319719  9.385105
#> 11  8.343616  8.343616  8.343616  8.343616  8.343616  8.363521  8.420443
#> 12  8.776408  8.776408  8.776408  8.776408  8.776408  8.783161  8.802474
#> 13 26.005904 26.005904 26.005904 26.005904 26.005904 26.040335 26.138798
#> 14 29.210755 29.210755 29.210755 29.210755 29.210755 29.209504 29.205928
#> 15 28.005512 28.005512 28.005512 28.005512 28.005512 28.049046 28.173541
#> 16 24.554134 24.554134 24.554134 24.554134 24.554134 24.608092 24.762394
#> 17 19.240108 19.240108 19.240108 19.240108 19.240108 19.221808 19.169475
#> 18 17.021365 17.021365 17.021365 17.021365 17.021365 16.949633 16.744502
#> 19 26.334607 26.334607 26.334607 26.334607 26.334607 26.283413 26.137018
#> 20 29.769549 29.769549 29.769549 29.769549 29.769549 29.717473 29.568555
#> 21 18.500527 18.500527 18.500527 18.500527 18.500527 18.400931 18.116120
#>       1.2338    1.1507    1.0731    1.0008    0.9333    0.8704    0.8118
#> 1  22.091627 22.009178 21.932285 21.860607 21.793728 21.924734 23.034797
#> 2  25.409838 25.433212 25.455011 25.475331 25.494290 25.704159 26.882381
#> 3  20.738180 20.849764 20.953828 21.050833 21.141346 21.111562 20.500838
#> 4  19.037606 18.970880 18.908651 18.850642 18.796516 18.521517 17.116986
#> 5  18.098730 17.945624 17.802836 17.669733 17.545539 17.155071 15.386069
#> 6  21.782058 22.155726 22.504209 22.829056 23.132159 23.504828 24.314884
#> 7  19.413895 19.490555 19.562048 19.628692 19.690876 19.641684 19.048465
#> 8  15.708296 15.688084 15.669234 15.651663 15.635268 15.533176 14.994149
#> 9  17.395939 17.413418 17.429719 17.444915 17.459093 17.370485 16.767468
#> 10  9.446076  9.502938  9.555968  9.605401  9.651525  9.857466 10.882690
#> 11  8.473521  8.523022  8.569187  8.612221  8.652374  8.874247 10.024175
#> 12  8.820482  8.837277  8.852940  8.867541  8.881164  9.039034  9.928374
#> 13 26.230612 26.316238 26.396093 26.470532 26.539988 26.828910 28.244643
#> 14 29.202594 29.199484 29.196584 29.193881 29.191359 29.248001 29.602354
#> 15 28.289629 28.397894 28.498862 28.592981 28.680800 28.947637 30.142502
#> 16 24.906278 25.040464 25.165606 25.282261 25.391107 25.278281 24.078400
#> 17 19.120676 19.075166 19.032723 18.993159 18.956243 18.713865 17.424599
#> 18 16.553221 16.374832 16.208466 16.053384 15.908683 15.521109 13.867191
#> 19 26.000508 25.873198 25.754468 25.643792 25.540523 25.484619 25.638149
#> 20 29.429691 29.300186 29.179410 29.066825 28.961777 28.803236 28.344900
#> 21 17.850540 17.602860 17.371872 17.156551 16.955643 16.936378 17.775986
#>      0.7571   0.7060   0.6585   0.6141   0.5727   0.5341   0.4981   0.4645
#> 1  23.45846 23.45654 23.45682 23.45699 23.45709 23.34032 23.09986 23.04466
#> 2  27.37301 27.39141 27.39168 27.39184 27.39194 27.72547 28.41521 28.57279
#> 3  20.29221 20.30546 20.30530 20.30519 20.30513 20.48308 20.85024 20.93435
#> 4  16.51776 16.48867 16.48836 16.48817 16.48806 16.71832 17.19305 17.30185
#> 5  14.60584 14.55647 14.55610 14.55587 14.55574 14.43618 14.18805 14.13155
#> 6  24.78427 24.86282 24.86292 24.86298 24.86301 24.42089 23.50808 23.29926
#> 7  18.83410 18.84113 18.84097 18.84088 18.84082 18.37674 17.41767 17.19843
#> 8  14.76617 14.75600 14.75588 14.75581 14.75577 14.85806 15.06902 15.11735
#> 9  16.52723 16.52324 16.52310 16.52301 16.52296 16.48956 16.42018 16.40441
#> 10 11.32310 11.34584 11.34606 11.34620 11.34628 11.44195 11.64026 11.68546
#> 11 10.51288 10.53576 10.53601 10.53616 10.53626 10.69302 11.01759 11.09165
#> 12 10.29840 10.31213 10.31233 10.31245 10.31253 10.48959 10.85595 10.93960
#> 13 28.85543 28.88814 28.88845 28.88863 28.88874 29.10395 29.54941 29.65109
#> 14 29.74617 29.74983 29.74992 29.74997 29.75000 29.65944 29.47262 29.42983
#> 15 30.67129 30.70553 30.70579 30.70594 30.70603 30.78350 30.94431 30.98092
#> 16 23.63710 23.64747 23.64717 23.64698 23.64687 23.57368 23.42167 23.38714
#> 17 16.88036 16.85656 16.85627 16.85610 16.85599 17.05651 17.46986 17.56461
#> 18 13.12470 13.07205 13.07171 13.07151 13.07138 13.06508 13.05099 13.04800
#> 19 25.65394 25.63232 25.63238 25.63242 25.63245 25.86213 26.33670 26.44518
#> 20 28.10957 28.08023 28.08015 28.08010 28.08008 28.12792 28.22645 28.24903
#> 21 18.02799 17.99239 17.99264 17.99280 17.99289 17.29463 15.85283 15.52284
#>      0.4332   0.4040   0.3768   0.3514   0.3277   0.3056   0.2850   0.2658
#> 1  23.04477 23.04485 23.04491 23.04496 23.04499 23.04501 23.04503 23.04504
#> 2  28.57321 28.57352 28.57375 28.57391 28.57403 28.57411 28.57417 28.57421
#> 3  20.93436 20.93436 20.93437 20.93437 20.93437 20.93437 20.93438 20.93438
#> 4  17.30179 17.30174 17.30171 17.30168 17.30166 17.30165 17.30164 17.30163
#> 5  14.13119 14.13093 14.13074 14.13061 14.13051 14.13044 14.13039 14.13035
#> 6  23.29904 23.29888 23.29876 23.29868 23.29862 23.29857 23.29854 23.29852
#> 7  17.19799 17.19768 17.19745 17.19729 17.19717 17.19708 17.19702 17.19698
#> 8  15.11734 15.11732 15.11732 15.11731 15.11731 15.11730 15.11730 15.11730
#> 9  16.40428 16.40419 16.40413 16.40408 16.40404 16.40402 16.40400 16.40399
#> 10 11.68569 11.68586 11.68598 11.68607 11.68613 11.68618 11.68621 11.68623
#> 11 11.09194 11.09216 11.09231 11.09242 11.09250 11.09256 11.09260 11.09263
#> 12 10.93987 10.94007 10.94021 10.94031 10.94038 10.94044 10.94048 10.94050
#> 13 29.65147 29.65174 29.65194 29.65208 29.65218 29.65225 29.65231 29.65235
#> 14 29.42983 29.42982 29.42982 29.42982 29.42982 29.42982 29.42982 29.42982
#> 15 30.98116 30.98133 30.98146 30.98155 30.98161 30.98166 30.98170 30.98172
#> 16 23.38687 23.38668 23.38654 23.38644 23.38636 23.38631 23.38627 23.38624
#> 17 17.56454 17.56449 17.56445 17.56442 17.56440 17.56439 17.56438 17.56437
#> 18 13.04774 13.04755 13.04742 13.04732 13.04725 13.04720 13.04716 13.04714
#> 19 26.44539 26.44553 26.44564 26.44572 26.44577 26.44581 26.44584 26.44586
#> 20 28.24901 28.24899 28.24897 28.24896 28.24896 28.24895 28.24895 28.24894
#> 21 15.52252 15.52229 15.52213 15.52201 15.52192 15.52186 15.52182 15.52178
#>      0.2479   0.2312   0.2156   0.2011   0.1875   0.1749   0.1631   0.1521
#> 1  23.04505 23.04505 23.04506 22.95341 22.95319 22.95304 22.38073 22.38027
#> 2  28.57425 28.57427 28.57429 28.36784 28.36751 28.36728 28.78053 28.78076
#> 3  20.93438 20.93438 20.93438 20.43851 20.43845 20.43843 20.92651 20.92666
#> 4  17.30163 17.30163 17.30162 17.15008 17.15018 17.15026 16.87171 16.87164
#> 5  14.13032 14.13030 14.13029 14.01401 14.01421 14.01438 13.98036 13.98035
#> 6  23.29850 23.29849 23.29848 23.34838 23.34852 23.34863 22.87754 22.87700
#> 7  17.19694 17.19692 17.19690 17.59504 17.59540 17.59565 18.18203 18.18238
#> 8  15.11730 15.11730 15.11730 14.93855 14.93855 14.93855 14.73881 14.73869
#> 9  16.40398 16.40397 16.40397 16.04606 16.04605 16.04607 15.57157 15.57115
#> 10 11.68625 11.68626 11.68627 11.58516 11.58497 11.58483 11.62355 11.62351
#> 11 11.09266 11.09267 11.09268 11.16653 11.16635 11.16622 11.32086 11.32099
#> 12 10.94052 10.94054 10.94055 11.33046 11.33042 11.33036 11.57558 11.57593
#> 13 29.65237 29.65239 29.65241 29.59880 29.59854 29.59835 29.74925 29.74932
#> 14 29.42982 29.42982 29.42982 30.25746 30.25770 30.25783 30.42370 30.42411
#> 15 30.98174 30.98175 30.98176 30.92267 30.92248 30.92235 30.82337 30.82324
#> 16 23.38623 23.38621 23.38620 23.34753 23.34779 23.34798 23.42785 23.42788
#> 17 17.56437 17.56437 17.56436 17.37809 17.37817 17.37824 17.02288 17.02273
#> 18 13.04712 13.04710 13.04709 13.75797 13.75840 13.75868 13.97965 13.98019
#> 19 26.44588 26.44589 26.44590 27.03451 27.03459 27.03460 26.41804 26.41804
#> 20 28.24894 28.24894 28.24894 27.72167 27.72149 27.72140 28.20516 28.20528
#> 21 15.52176 15.52174 15.52173 15.04728 15.04702 15.04689 15.12032 15.11989
#>      0.1419   0.1323   0.1234   0.1151   0.1073   0.1001   0.0933   0.0870
#> 1  22.37989 22.37958 22.37931 21.52456 21.52435 21.52415 21.52396 20.75153
#> 2  28.78095 28.78111 28.78124 28.46253 28.46256 28.46259 28.46263 28.66853
#> 3  20.92678 20.92688 20.92697 21.11619 21.11631 21.11642 21.11652 20.95198
#> 4  16.87158 16.87154 16.87150 17.78312 17.78334 17.78354 17.78373 17.94250
#> 5  13.98034 13.98033 13.98033 13.15235 13.15217 13.15200 13.15185 13.34437
#> 6  22.87655 22.87618 22.87588 23.09146 23.09136 23.09126 23.09117 23.47895
#> 7  18.18267 18.18291 18.18311 19.08445 19.08456 19.08467 19.08477 18.43712
#> 8  14.73858 14.73849 14.73842 14.98570 14.98576 14.98582 14.98587 14.83473
#> 9  15.57079 15.57050 15.57025 15.93839 15.93844 15.93849 15.93854 15.91766
#> 10 11.62348 11.62345 11.62343 11.26575 11.26569 11.26565 11.26560 11.06524
#> 11 11.32109 11.32118 11.32125 11.04895 11.04892 11.04890 11.04888 10.99058
#> 12 11.57621 11.57645 11.57664 11.51323 11.51325 11.51327 11.51329 11.72501
#> 13 29.74938 29.74943 29.74947 29.98681 29.98691 29.98701 29.98710 30.11492
#> 14 30.42444 30.42473 30.42496 30.77613 30.77618 30.77623 30.77627 30.65832
#> 15 30.82314 30.82305 30.82298 31.32453 31.32465 31.32477 31.32487 31.61955
#> 16 23.42790 23.42792 23.42794 21.27666 21.27619 21.27576 21.27536 21.31487
#> 17 17.02261 17.02251 17.02243 18.01187 18.01209 18.01230 18.01249 17.87208
#> 18 13.98064 13.98102 13.98133 13.83260 13.83256 13.83253 13.83250 14.22590
#> 19 26.41804 26.41804 26.41803 26.24247 26.24239 26.24232 26.24224 26.22263
#> 20 28.20538 28.20546 28.20553 28.28054 28.28061 28.28067 28.28073 28.15604
#> 21 15.11954 15.11924 15.11900 15.30173 15.30168 15.30165 15.30161 15.70752
#>      0.0812   0.0757   0.0706   0.0658   0.0614   0.0573   0.0534   0.0498
#> 1  20.75140 20.75127 20.75114 20.75101 20.75088 20.75075 20.75063 20.75050
#> 2  28.66853 28.66854 28.66855 28.66856 28.66858 28.66860 28.66863 28.66866
#> 3  20.95212 20.95225 20.95236 20.95246 20.95255 20.95263 20.95269 20.95275
#> 4  17.94271 17.94292 17.94311 17.94328 17.94345 17.94360 17.94375 17.94389
#> 5  13.34428 13.34419 13.34412 13.34406 13.34400 13.34395 13.34390 13.34386
#> 6  23.47885 23.47876 23.47868 23.47861 23.47856 23.47851 23.47847 23.47844
#> 7  18.43720 18.43726 18.43729 18.43730 18.43729 18.43726 18.43723 18.43718
#> 8  14.83481 14.83489 14.83495 14.83502 14.83507 14.83512 14.83516 14.83519
#> 9  15.91779 15.91791 15.91803 15.91813 15.91823 15.91832 15.91841 15.91848
#> 10 11.06518 11.06511 11.06505 11.06500 11.06494 11.06489 11.06484 11.06479
#> 11 10.99050 10.99043 10.99037 10.99031 10.99026 10.99021 10.99017 10.99013
#> 12 11.72494 11.72488 11.72483 11.72479 11.72475 11.72472 11.72470 11.72469
#> 13 30.11500 30.11508 30.11515 30.11523 30.11531 30.11538 30.11545 30.11552
#> 14 30.65834 30.65836 30.65837 30.65839 30.65840 30.65841 30.65842 30.65842
#> 15 31.61968 31.61981 31.61995 31.62008 31.62020 31.62033 31.62045 31.62057
#> 16 21.31448 21.31412 21.31379 21.31348 21.31320 21.31293 21.31269 21.31247
#> 17 17.87231 17.87252 17.87272 17.87289 17.87305 17.87320 17.87333 17.87345
#> 18 14.22583 14.22577 14.22571 14.22567 14.22564 14.22561 14.22559 14.22558
#> 19 26.22238 26.22216 26.22195 26.22175 26.22157 26.22140 26.22125 26.22111
#> 20 28.15608 28.15612 28.15615 28.15616 28.15617 28.15618 28.15618 28.15618
#> 21 15.70759 15.70767 15.70775 15.70783 15.70791 15.70799 15.70807 15.70815
#>      0.0465   0.0433   0.0404   0.0377   0.0351   0.0328   0.0306   0.0285
#> 1  20.75038 20.75026 20.75014 20.75002 20.74990 20.74979 20.74967 20.74956
#> 2  28.66869 28.66872 28.66876 28.66879 28.66883 28.66887 28.66890 28.66894
#> 3  20.95280 20.95284 20.95287 20.95290 20.95292 20.95294 20.95295 20.95296
#> 4  17.94401 17.94413 17.94424 17.94434 17.94443 17.94452 17.94460 17.94468
#> 5  13.34382 13.34379 13.34376 13.34373 13.34371 13.34369 13.34366 13.34364
#> 6  23.47842 23.47840 23.47838 23.47837 23.47837 23.47837 23.47837 23.47838
#> 7  18.43712 18.43706 18.43699 18.43692 18.43685 18.43677 18.43670 18.43663
#> 8  14.83522 14.83525 14.83527 14.83529 14.83530 14.83531 14.83532 14.83533
#> 9  15.91855 15.91861 15.91867 15.91872 15.91876 15.91880 15.91883 15.91886
#> 10 11.06475 11.06470 11.06466 11.06462 11.06459 11.06455 11.06452 11.06449
#> 11 10.99009 10.99006 10.99003 10.99001 10.98998 10.98997 10.98995 10.98994
#> 12 11.72468 11.72467 11.72468 11.72468 11.72469 11.72471 11.72472 11.72474
#> 13 30.11559 30.11566 30.11572 30.11578 30.11584 30.11590 30.11595 30.11600
#> 14 30.65843 30.65843 30.65844 30.65844 30.65844 30.65845 30.65845 30.65845
#> 15 31.62068 31.62079 31.62090 31.62100 31.62109 31.62118 31.62126 31.62134
#> 16 21.31226 21.31208 21.31190 21.31174 21.31159 21.31146 21.31133 21.31121
#> 17 17.87356 17.87365 17.87374 17.87381 17.87388 17.87394 17.87399 17.87404
#> 18 14.22557 14.22557 14.22557 14.22558 14.22559 14.22561 14.22562 14.22564
#> 19 26.22099 26.22088 26.22078 26.22070 26.22062 26.22056 26.22050 26.22046
#> 20 28.15617 28.15616 28.15615 28.15614 28.15613 28.15612 28.15610 28.15609
#> 21 15.70822 15.70829 15.70836 15.70842 15.70848 15.70853 15.70858 15.70862
#>      0.0266   0.0248   0.0231   0.0216   0.0201   0.0188   0.0175   0.0163
#> 1  20.74926 20.75217 20.75250 20.75282 20.75312 20.75341 20.75369 20.75395
#> 2  28.66918 28.52678 28.52677 28.52676 28.52675 28.52674 28.52673 28.52672
#> 3  20.95278 21.06862 21.06851 21.06840 21.06830 21.06820 21.06811 21.06802
#> 4  17.94500 18.13475 18.13448 18.13422 18.13397 18.13373 18.13350 18.13328
#> 5  13.34330 13.45890 13.45904 13.45916 13.45926 13.45936 13.45945 13.45953
#> 6  23.47830 23.53701 23.53703 23.53704 23.53705 23.53706 23.53707 23.53707
#> 7  18.43589 18.31438 18.31434 18.31430 18.31428 18.31427 18.31426 18.31427
#> 8  14.83450 14.52712 14.52715 14.52718 14.52722 14.52726 14.52730 14.52735
#> 9  15.91811 15.75904 15.75903 15.75903 15.75903 15.75903 15.75904 15.75904
#> 10 11.06507 11.29656 11.29660 11.29663 11.29666 11.29668 11.29670 11.29671
#> 11 10.99048 11.05918 11.05920 11.05922 11.05924 11.05925 11.05926 11.05928
#> 12 11.72536 11.65791 11.65787 11.65784 11.65781 11.65778 11.65775 11.65772
#> 13 30.11636 30.00377 30.00369 30.00361 30.00354 30.00347 30.00341 30.00335
#> 14 30.65942 30.83135 30.83126 30.83117 30.83109 30.83101 30.83094 30.83088
#> 15 31.62187 31.63887 31.63872 31.63858 31.63844 31.63831 31.63819 31.63807
#> 16 21.31039 21.24636 21.24682 21.24726 21.24768 21.24807 21.24843 21.24878
#> 17 17.87386 17.89441 17.89421 17.89402 17.89384 17.89367 17.89352 17.89337
#> 18 14.22582 14.14514 14.14512 14.14510 14.14508 14.14506 14.14504 14.14503
#> 19 26.22104 26.21186 26.21193 26.21199 26.21205 26.21211 26.21216 26.21221
#> 20 28.15580 28.23299 28.23292 28.23285 28.23279 28.23273 28.23268 28.23262
#> 21 15.70818 15.70283 15.70282 15.70281 15.70280 15.70279 15.70277 15.70276
#>      0.0152   0.0142   0.0132   0.0123   0.0115   0.0107   0.0100   0.0093
#> 1  20.75420 20.75444 20.75466 20.75488 20.75508 20.75527 20.75545 20.75563
#> 2  28.52671 28.52670 28.52669 28.52668 28.52667 28.52667 28.52666 28.52665
#> 3  21.06794 21.06786 21.06778 21.06771 21.06764 21.06758 21.06752 21.06746
#> 4  18.13307 18.13286 18.13267 18.13248 18.13231 18.13214 18.13198 18.13183
#> 5  13.45961 13.45967 13.45973 13.45978 13.45982 13.45986 13.45990 13.45993
#> 6  23.53708 23.53708 23.53707 23.53707 23.53707 23.53706 23.53706 23.53705
#> 7  18.31429 18.31431 18.31434 18.31438 18.31442 18.31447 18.31452 18.31458
#> 8  14.52739 14.52744 14.52749 14.52754 14.52760 14.52765 14.52770 14.52775
#> 9  15.75905 15.75905 15.75906 15.75907 15.75908 15.75909 15.75910 15.75911
#> 10 11.29672 11.29673 11.29673 11.29674 11.29673 11.29673 11.29673 11.29672
#> 11 11.05928 11.05929 11.05930 11.05930 11.05931 11.05931 11.05931 11.05932
#> 12 11.65770 11.65767 11.65765 11.65763 11.65761 11.65760 11.65758 11.65757
#> 13 30.00329 30.00324 30.00319 30.00315 30.00310 30.00307 30.00303 30.00300
#> 14 30.83081 30.83075 30.83070 30.83065 30.83060 30.83056 30.83051 30.83048
#> 15 31.63795 31.63784 31.63773 31.63763 31.63753 31.63744 31.63735 31.63727
#> 16 21.24910 21.24941 21.24969 21.24996 21.25021 21.25044 21.25066 21.25086
#> 17 17.89323 17.89309 17.89297 17.89286 17.89275 17.89265 17.89255 17.89247
#> 18 14.14501 14.14500 14.14499 14.14498 14.14497 14.14496 14.14495 14.14494
#> 19 26.21226 26.21231 26.21235 26.21240 26.21244 26.21247 26.21251 26.21254
#> 20 28.23257 28.23252 28.23247 28.23243 28.23239 28.23235 28.23231 28.23228
#> 21 15.70274 15.70272 15.70270 15.70268 15.70266 15.70264 15.70262 15.70260
#>      0.0087   0.0081   0.0076   0.0071   0.0066   0.0061
#> 1  20.75579 20.75594 20.75609 20.75622 20.75635 20.75647
#> 2  28.52664 28.52663 28.52663 28.52662 28.52661 28.52660
#> 3  21.06741 21.06735 21.06731 21.06726 21.06722 21.06718
#> 4  18.13168 18.13155 18.13142 18.13129 18.13118 18.13107
#> 5  13.45995 13.45997 13.45999 13.46000 13.46001 13.46002
#> 6  23.53704 23.53703 23.53702 23.53701 23.53700 23.53699
#> 7  18.31463 18.31469 18.31475 18.31482 18.31488 18.31495
#> 8  14.52780 14.52785 14.52790 14.52795 14.52800 14.52804
#> 9  15.75912 15.75913 15.75914 15.75915 15.75916 15.75917
#> 10 11.29672 11.29671 11.29670 11.29669 11.29668 11.29667
#> 11 11.05932 11.05932 11.05932 11.05932 11.05932 11.05931
#> 12 11.65755 11.65754 11.65753 11.65752 11.65750 11.65749
#> 13 30.00296 30.00294 30.00291 30.00288 30.00286 30.00284
#> 14 30.83044 30.83041 30.83038 30.83035 30.83033 30.83030
#> 15 31.63719 31.63711 31.63704 31.63697 31.63690 31.63684
#> 16 21.25105 21.25122 21.25139 21.25154 21.25167 21.25180
#> 17 17.89239 17.89231 17.89224 17.89218 17.89212 17.89206
#> 18 14.14493 14.14492 14.14492 14.14491 14.14490 14.14490
#> 19 26.21257 26.21260 26.21263 26.21265 26.21268 26.21270
#> 20 28.23224 28.23221 28.23218 28.23215 28.23213 28.23210
#> 21 15.70258 15.70256 15.70254 15.70252 15.70250 15.70248
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 22.8 21.4 18.7 14.3 22.8 17.8 16.4 15.2 10.4 10.4 14.7 32.4 30.4 33.9
#> [16] 21.5 15.2 13.3 26.0 30.4 15.0
#> 
#> $X
#>               am       carb        cyl       disp       drat       gear
#>  [1,]  1.2747549  0.7210872 -0.1034175 -0.5911085  0.5445033  0.5275044
#>  [2,]  1.2747549 -1.0960525 -1.1893017 -0.9861456  0.4495318  0.5275044
#>  [3,] -0.7844645 -1.0960525 -0.1034175  0.1533843 -1.0130293 -0.8571946
#>  [4,] -0.7844645 -0.4903393  0.9824666  0.9282647 -0.8800692 -0.8571946
#>  [5,] -0.7844645  0.7210872  0.9824666  0.9282647 -0.7661034 -0.8571946
#>  [6,] -0.7844645 -0.4903393 -1.1893017 -0.7369684  0.5824919  0.5275044
#>  [7,] -0.7844645  0.7210872 -0.1034175 -0.5333724  0.5824919  0.5275044
#>  [8,] -0.7844645  0.1153739  0.9824666  0.2886086 -1.0320236 -0.8571946
#>  [9,] -0.7844645  0.1153739  0.9824666  0.2886086 -1.0320236 -0.8571946
#> [10,] -0.7844645  0.7210872  0.9824666  1.7791137 -1.2979438 -0.8571946
#> [11,] -0.7844645  0.7210872  0.9824666  1.6879513 -1.1649837 -0.8571946
#> [12,] -0.7844645  0.7210872  0.9824666  1.5360140 -0.7281148 -0.8571946
#> [13,]  1.2747549 -1.0960525 -1.1893017 -1.2087338  0.8864006  0.5275044
#> [14,]  1.2747549 -0.4903393 -1.1893017 -1.2315244  2.5009161  0.5275044
#> [15,]  1.2747549 -1.0960525 -1.1893017 -1.2664700  1.1523208  0.5275044
#> [16,] -0.7844645 -1.0960525 -1.1893017 -0.8942235  0.1646173 -0.8571946
#> [17,] -0.7844645 -0.4903393  0.9824666  0.5028402 -0.8800692 -0.8571946
#> [18,] -0.7844645  0.7210872  0.9824666  0.8522960  0.2216002 -0.8571946
#> [19,]  1.2747549 -0.4903393 -1.1893017 -0.8927041  1.5512011  1.9122034
#> [20,]  1.2747549 -0.4903393 -1.1893017 -1.0841452  0.2975774  1.9122034
#> [21,]  1.2747549  3.1439401  0.9824666  0.4800496 -0.1392915  1.9122034
#>                hp        qsec         vs          wt
#>  [1,] -0.56895435 -0.52908161 -0.8660254 -0.34482967
#>  [2,] -0.80377135  0.36585430  1.1547005 -0.83996208
#>  [3,] -0.56895435  0.83302210  1.1547005 -0.04150531
#>  [4,]  0.32887535 -0.52908161 -0.8660254  0.15922405
#>  [5,]  1.29576888 -1.19324788 -0.8660254  0.27520101
#>  [6,] -0.77614582  2.78049270  1.1547005 -0.09949379
#>  [7,] -0.38938841  0.52908161  1.1547005  0.15922405
#>  [8,]  0.39793917 -0.31519755 -0.8660254  0.72126624
#>  [9,]  0.39793917  0.02251411 -0.8660254  0.46254841
#> [10,]  0.74325829  0.01125706 -0.8660254  1.77398020
#> [11,]  0.88138594 -0.07879939 -0.8660254  1.92921090
#> [12,]  1.08857741 -0.30394050 -0.8660254  1.85873259
#> [13,] -1.17671600  0.84990769  1.1547005 -0.94701773
#> [14,] -1.37009471  0.31519755  1.1547005 -1.46891406
#> [15,] -1.19052876  1.09193438  1.1547005 -1.27264536
#> [16,] -0.74852030  1.15384818  1.1547005 -0.71060316
#> [17,] -0.01644377 -0.37148283 -0.8660254  0.15476340
#> [18,]  1.29576888 -1.43527457 -0.8660254  0.51607624
#> [19,] -0.83139688 -0.70919449 -0.8660254 -1.00054556
#> [20,] -0.52751606 -0.59662394  1.1547005 -1.55991137
#> [21,]  2.53891770 -1.89118532 -0.8660254  0.27520101
#> attr(,"center")
#>  [1]   0.3809524   2.8095238   6.1904762 237.8095238   3.6133333   3.6190476
#>  [7] 151.1904762  17.9600000   0.4285714   3.2615238
#> attr(,"scale")
#>  [1]   0.4856209   1.6509463   1.8418171 131.6332239   0.5264737   0.7221786
#>  [7]  72.3968028   1.7766635   0.4948717   1.1209123
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 42
#> 
#> $lambda.min
#> [1] 0.351398
#> 
#> $null.dev
#> [1] 47.60848
#> 
#> $Bias
#> [1] 7.529802
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 8.761154