Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 46.03911 43.93136 40.79915 35.58322 31.06258 27.35136 24.32925 21.89263
#>   [9] 19.95232 18.43160 17.26458 16.39471 15.77349 15.35941 15.11699 15.01549
#>  [17] 14.70257 14.28668 13.99227 14.08175 14.19823 14.32579 14.44544 14.55633
#>  [25] 14.46553 14.15075 14.34507 14.43025 15.03942 15.34210 15.41399 15.41003
#>  [33] 15.44447 16.35547 16.45354 16.85105 15.03655 14.61598 14.62124 14.10344
#>  [41] 15.24581 14.22067 16.51772 15.51720 17.50630 23.08734 24.59686 24.21299
#>  [49] 23.25143 25.62150 26.82139 26.65319 27.00366 28.13365 28.11623 26.07027
#>  [57] 26.08796 25.40394 24.53703 24.63364 24.63462 23.90016 23.70519 23.55910
#>  [65] 23.28979 23.28999 23.29018 23.29035 22.17040 22.17052 22.01143 22.01149
#>  [73] 22.01154 22.01157 22.18585 22.18547 22.18509 22.18471 22.18434 22.18397
#>  [81] 22.18359 22.18323 22.18286 22.18250 22.18214 23.28488 23.28522 23.28555
#>  [89] 23.28589 23.28623 23.28657 23.28690 23.28724 23.28758 23.28791 23.28825
#>  [97] 23.28864 23.27442 23.27464 23.27487
#> 
#> $cvse
#>   [1] 11.915976 11.947705 11.485340 10.260118  9.139556  8.192704  7.387231
#>   [8]  6.698241  6.107394  5.601921  5.173579  4.817544  4.531246  4.313142
#>  [15]  4.161528  4.073606  3.962819  3.820879  3.722101  3.710179  3.696973
#>  [22]  3.686375  3.682312  3.686170  3.681043  3.781331  3.871543  3.853729
#>  [29]  3.765354  3.755547  3.775608  3.697349  3.692788  4.076217  4.077399
#>  [36]  4.042266  3.860045  3.658973  3.634046  3.429755  3.540407  3.549056
#>  [43]  3.745816  3.684949  4.506466  7.536579  7.333937  7.311662  7.278074
#>  [50]  7.446249  8.484065  8.498124  8.468893  8.656878  8.656671  7.494375
#>  [57]  7.490691  7.426776  7.255889  7.241270  7.241573  6.080392  6.069541
#>  [64]  6.063994  6.096674  6.096466  6.096262  6.096062  6.119277  6.119097
#>  [71]  6.127986  6.127817  6.127651  6.127488  5.940009  5.939516  5.939034
#>  [78]  5.938564  5.938104  5.937654  5.937215  5.936786  5.936367  5.935957
#>  [85]  5.935557  6.057690  6.057473  6.057264  6.057061  6.056865  6.056675
#>  [92]  6.056491  6.056312  6.056140  6.055973  6.055812  6.055650  6.056509
#>  [99]  6.056373  6.056242
#> 
#> $fold
#>  [1]  3  6 10  4  6 10  7  3  5  1  9  9  2  2  5  8  1  7  1  8  4
#> 
#> $lambda
#>   [1] 5.390408923 5.027113403 4.688302785 4.372326868 4.077646671 3.802826933
#>   [7] 3.546529125 3.307504932 3.084590169 2.876699116 2.682819223 2.502006187
#>  [13] 2.333379344 2.176117385 2.029454357 1.892675926 1.765115903 1.646152999
#>  [19] 1.535207796 1.431739929 1.335245449 1.245254375 1.161328398 1.083058751
#>  [25] 1.010064216 0.941989268 0.878502344 0.819294226 0.764076538 0.712580336
#>  [31] 0.664554807 0.619766037 0.577995880 0.539040892 0.502711340 0.468830278
#>  [37] 0.437232687 0.407764667 0.380282693 0.354652913 0.330750493 0.308459017
#>  [43] 0.287669912 0.268281922 0.250200619 0.233337934 0.217611739 0.202945436
#>  [49] 0.189267593 0.176511590 0.164615300 0.153520780 0.143173993 0.133524545
#>  [55] 0.124525437 0.116132840 0.108305875 0.101006422 0.094198927 0.087850234
#>  [61] 0.081929423 0.076407654 0.071258034 0.066455481 0.061976604 0.057799588
#>  [67] 0.053904089 0.050271134 0.046883028 0.043723269 0.040776467 0.038028269
#>  [73] 0.035465291 0.033075049 0.030845902 0.028766991 0.026828192 0.025020062
#>  [79] 0.023333793 0.021761174 0.020294544 0.018926759 0.017651159 0.016461530
#>  [85] 0.015352078 0.014317399 0.013352454 0.012452544 0.011613284 0.010830588
#>  [91] 0.010100642 0.009419893 0.008785023 0.008192942 0.007640765 0.007125803
#>  [97] 0.006645548 0.006197660 0.005779959 0.005390409
#> 
#> $fit
#> $beta
#>               5.3904     5.0271    4.6883    4.3723    4.0776    3.8028
#> (Intercept) 21.70952 23.3000693 24.783417 26.166793 27.456933 28.660122
#> am           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.00000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.00000 -0.5225755 -1.009931 -1.464441 -1.888318 -2.283627
#>                3.5465    3.3075    3.0846    2.8767   2.6828    2.5020
#> (Intercept) 29.782221 30.828694 31.804638 32.714806 33.56363 34.355251
#> am           0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> wt          -2.652293 -2.996113 -3.316761 -3.615798 -3.89468 -4.154767
#>                2.3334    2.1761   2.0295    1.8927    1.7651    1.6462
#> (Intercept) 35.093517 35.782026 36.42413 37.022962 37.442704 37.442704
#> am           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.00000  0.000000  0.000000  0.000000
#> wt          -4.397325 -4.623536 -4.83450 -5.031247 -5.169153 -5.169153
#>                1.5352    1.4317    1.3352    1.2453      1.1613     1.0831
#> (Intercept) 37.442704 37.442704 37.442704 37.442704 37.23608420 35.5211018
#> am           0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.01107312  0.1029823
#> vs           0.000000  0.000000  0.000000  0.000000  0.00000000  0.0000000
#> wt          -5.169153 -5.169153 -5.169153 -5.169153 -5.16697218 -5.1488680
#>                1.0101     0.9420     0.8785     0.8193     0.7641     0.7126
#> (Intercept) 33.921874 32.4304282 31.0395012 29.7423180 28.5325606 27.4043368
#> am           0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.188688  0.2686173  0.3431598  0.4126783  0.4775115  0.5379751
#> vs           0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.131986 -5.1162415 -5.1015583 -5.0878646 -5.0750939 -5.0631838
#>                 0.6646     0.6198     0.5780     0.5390    0.5027    0.4688
#> (Intercept) 25.6579234 22.2715501 19.1192262 16.1754222 19.200456  9.076974
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000 -1.139124 -1.886648
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> drat         0.1609736  0.7174442  1.2351657  1.7188365  3.946004  3.322211
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  2.710817
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> qsec         0.5896044  0.6260199  0.6599938  0.6916694  0.000000  0.000000
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> wt          -4.9904759 -4.7671144 -4.5592879 -4.3651428 -2.986716 -1.625003
#>                0.4372    0.4078    0.3803    0.3547    0.3308    0.3085
#> (Intercept)  9.074575  9.072679  9.071181  9.069998  9.069063  9.068324
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb        -1.886870 -1.887045 -1.887183 -1.887292 -1.887378 -1.887446
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         3.322147  3.322097  3.322057  3.322026  3.322001  3.321982
#> gear         2.711386  2.711835  2.712190  2.712471  2.712693  2.712868
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt          -1.624654 -1.624378 -1.624160 -1.623988 -1.623852 -1.623745
#>                 0.2877     0.2683     0.2502     0.2333     0.2176     0.2029
#> (Intercept) -1.8127431 -1.8159447 -1.8183781 -1.8202236 -1.8216205 -1.8226762
#> am           3.0624963  3.0627677  3.0628952  3.0629321  3.0629149  3.0628678
#> carb        -1.4726034 -1.4728153 -1.4730060 -1.4731734 -1.4733178 -1.4734403
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         1.8248945  1.8250031  1.8251074  1.8252016  1.8252833  1.8253525
#> gear         2.3696259  2.3700912  2.3705077  2.3708723  2.3711860  2.3714522
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.7918344  0.7918376  0.7918253  0.7918047  0.7917806  0.7917561
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -1.3415708 -1.3411080 -1.3407348 -1.3404352 -1.3401957 -1.3400050
#>                 0.1893     0.1765     0.1646     0.1535       0.1432
#> (Intercept) -1.8234729 -1.8240734 -36.557472 -36.564385 -51.07143325
#> am           3.0628065  3.0627408   3.677885   3.678161   3.55454460
#> carb        -1.4735431 -1.4736284  -1.791724  -1.791787  -1.88677431
#> cyl          0.0000000  0.0000000   1.620171   1.620489   1.76801295
#> disp         0.0000000  0.0000000   0.000000   0.000000   0.00000000
#> drat         1.8254100  1.8254570   2.641641   2.641615   3.23896785
#> gear         2.3716753  2.3718606   4.293346   4.293766   4.55641444
#> hp           0.0000000  0.0000000   0.000000   0.000000   0.03216199
#> qsec         0.7917328  0.7917115   1.732790   1.733004   2.28316512
#> vs           0.0000000  0.0000000   0.000000   0.000000   0.00000000
#> wt          -1.3398536 -1.3397338  -1.743565  -1.743636  -2.75807475
#>                   0.1335       0.1245       0.1161       0.1083       0.1010
#> (Intercept) -51.08119466 -51.09028989 -51.09876409 -58.94397650 -58.93694518
#> am            3.55470553   3.55485455   3.55499254   3.63121916   3.63106697
#> carb         -1.88687354  -1.88696614  -1.88705256  -1.07289821  -1.07279650
#> cyl           1.76834396   1.76865193   1.76893851   1.40674291   1.40641216
#> disp          0.00000000   0.00000000   0.00000000   0.02547054   0.02547017
#> drat          3.23914199   3.23930522   3.23945818   2.87297907   2.87287771
#> gear          4.55689071   4.55733420   4.55774720   4.73219150   4.73167440
#> hp            0.03216942   0.03217637   0.03218287   0.02202613   0.02202815
#> qsec          2.28347377   2.28376131   2.28402916   2.90937203   2.90919815
#> vs            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> wt           -2.75835963  -2.75862553  -2.75887365  -4.92206893  -4.92208358
#>                   0.0942       0.0879       0.0819       0.0764       0.0713
#> (Intercept) -58.93023221 -58.92382629 -58.91771637 -68.76222609 -68.76585976
#> am            3.63092068   3.63078014   3.63064519   3.33846529   3.33842610
#> carb         -1.07270650  -1.07262743  -1.07255855  -1.03064243  -1.03047774
#> cyl           1.40610226   1.40581219   1.40554094   1.41788572   1.41779519
#> disp          0.02546966   0.02546903   0.02546829   0.02547120   0.02547477
#> drat          2.87278647   2.87270473   2.87263187   2.97001748   2.96994096
#> gear          4.73118676   4.73072723   4.73029450   5.06785095   5.06785953
#> hp            0.02203000   0.02203171   0.02203327   0.02856997   0.02857378
#> qsec          2.90902794   2.90886146   2.90869879   3.42825613   3.42851650
#> vs            0.00000000   0.00000000   0.00000000  -1.29318586  -1.29348972
#> wt           -4.92207908  -4.92205703  -4.92201897  -5.35850937  -5.35907330
#>                   0.0665       0.0620       0.0578       0.0539       0.0503
#> (Intercept) -68.76934862 -68.77269945 -68.77591866 -68.77901233 -68.78198622
#> am            3.33838656   3.33834676   3.33830679   3.33826675   3.33822671
#> carb         -1.03031827  -1.03016391  -1.03001457  -1.02987014  -1.02973051
#> cyl           1.41770603   1.41761836   1.41753227   1.41744785   1.41736517
#> disp          0.02547821   0.02548153   0.02548472   0.02548778   0.02549073
#> drat          2.96986704   2.96979568   2.96972686   2.96966053   2.96959666
#> gear          5.06786557   5.06786937   5.06787119   5.06787128   5.06786984
#> hp            0.02857749   0.02858111   0.02858463   0.02858805   0.02859138
#> qsec          3.42876785   3.42901050   3.42924473   3.42947084   3.42968909
#> vs           -1.29378562  -1.29407368  -1.29435405  -1.29462685  -1.29489224
#> wt           -5.35961824  -5.36014473  -5.36065328  -5.36114440  -5.36161858
#>                   0.0469       0.0437       0.0408       0.0380       0.0355
#> (Intercept) -68.78484580 -68.78759625 -68.79024248 -68.79278914 -68.79524064
#> am            3.33818675   3.33814693   3.33810733   3.33806800   3.33802900
#> carb         -1.02959558  -1.02946524  -1.02933938  -1.02921788  -1.02910064
#> cyl           1.41728428   1.41720525   1.41712812   1.41705290   1.41697965
#> disp          0.02549357   0.02549630   0.02549892   0.02550145   0.02550387
#> drat          2.96953520   2.96947610   2.96941932   2.96936479   2.96931248
#> gear          5.06786709   5.06786321   5.06785838   5.06785274   5.06784645
#> hp            0.02859462   0.02859777   0.02860083   0.02860381   0.02860670
#> qsec          3.42989975   3.43010309   3.43029934   3.43048875   3.43067155
#> vs           -1.29515033  -1.29540126  -1.29564518  -1.29588222  -1.29611251
#> wt           -5.36207632  -5.36251810  -5.36294439  -5.36335567  -5.36375238
#>                   0.0331       0.0308       0.0288       0.0268       0.0250
#> (Intercept) -68.79760117 -68.79987471 -68.80206501 -68.80417565 -68.80621002
#> am            3.33799038   3.33795219   3.33791446   3.33787724   3.33784056
#> carb         -1.02898755  -1.02887849  -1.02877336  -1.02867204  -1.02857443
#> cyl           1.41690835   1.41683904   1.41677172   1.41670638   1.41664301
#> disp          0.02550620   0.02550843   0.02551058   0.02551264   0.02551462
#> drat          2.96926231   2.96921424   2.96916820   2.96912415   2.96908201
#> gear          5.06783962   5.06783239   5.06782484   5.06781708   5.06780919
#> hp            0.02860952   0.02861225   0.02861490   0.02861748   0.02861998
#> qsec          3.43084796   3.43101820   3.43118249   3.43134101   3.43149398
#> vs           -1.29633619  -1.29655339  -1.29676425  -1.29696890  -1.29716748
#> wt           -5.36413497  -5.36450389  -5.36485955  -5.36520238  -5.36553278
#>                   0.0233       0.0218       0.0203       0.0189       0.0177
#> (Intercept) -68.80817134 -68.81006266 -68.81188689 -68.81364678 -68.81534494
#> am            3.33780445   3.33776894   3.33773405   3.33769981   3.33766622
#> carb         -1.02848042  -1.02838991  -1.02830278  -1.02821893  -1.02813827
#> cyl           1.41658162   1.41652217   1.41646465   1.41640905   1.41635532
#> disp          0.02551652   0.02551834   0.02552009   0.02552176   0.02552337
#> drat          2.96904174   2.96900327   2.96896654   2.96893150   2.96889808
#> gear          5.06780125   5.06779332   5.06778547   5.06777774   5.06777018
#> hp            0.02862240   0.02862476   0.02862704   0.02862926   0.02863141
#> qsec          3.43164157   3.43178398   3.43192137   3.43205392   3.43218179
#> vs           -1.29736012  -1.29754695  -1.29772810  -1.29790370  -1.29807389
#> wt           -5.36585115  -5.36615789  -5.36645338  -5.36673798  -5.36701205
#>                   0.0165       0.0154       0.0143       0.0134       0.0125
#> (Intercept) -68.81698388 -68.81856595 -68.82009340 -68.82156838 -68.82299293
#> am            3.33763332   3.33760110   3.33756959   3.33753879   3.33750870
#> carb         -1.02806069  -1.02798609  -1.02791437  -1.02784545  -1.02777922
#> cyl           1.41630346   1.41625341   1.41620516   1.41615866   1.41611388
#> disp          0.02552492   0.02552640   0.02552781   0.02552917   0.02553048
#> drat          2.96886624   2.96883591   2.96880704   2.96877957   2.96875345
#> gear          5.06776283   5.06775573   5.06774890   5.06774237   5.06773616
#> hp            0.02863349   0.02863551   0.02863747   0.02863936   0.02864120
#> qsec          3.43230515   3.43242414   3.43253893   3.43264964   3.43275643
#> vs           -1.29823879  -1.29839852  -1.29855323  -1.29870302  -1.29884803
#> wt           -5.36727596  -5.36753004  -5.36777462  -5.36801003  -5.36823660
#>                   0.0116       0.0108       0.0101       0.0094       0.0088
#> (Intercept) -68.82436899 -68.82569840 -68.82698293 -68.82822426 -68.82942399
#> am            3.33747933   3.33745068   3.33742276   3.33739556   3.33736908
#> carb         -1.02771560  -1.02765449  -1.02759581  -1.02753948  -1.02748541
#> cyl           1.41607078   1.41602933   1.41598947   1.41595118   1.41591440
#> disp          0.02553173   0.02553293   0.02553407   0.02553517   0.02553623
#> drat          2.96872863   2.96870505   2.96868267   2.96866143   2.96864129
#> gear          5.06773028   5.06772475   5.06771958   5.06771476   5.06771031
#> hp            0.02864298   0.02864470   0.02864636   0.02864798   0.02864954
#> qsec          3.43285943   3.43295876   3.43305456   3.43314694   3.43323603
#> vs           -1.29898837  -1.29912417  -1.29925555  -1.29938261  -1.29950549
#> wt           -5.36845461  -5.36866438  -5.36886619  -5.36906032  -5.36924705
#>                   0.0082       0.0076       0.0071       0.0066       0.0062
#> (Intercept) -68.83058365 -68.83170470 -68.83278854 -68.83383649 -68.83484985
#> am            3.33734332   3.33731827   3.33729393   3.33727029   3.33724734
#> carb         -1.02743351  -1.02738372  -1.02733595  -1.02729014  -1.02724620
#> cyl           1.41587911   1.41584525   1.41581278   1.41578166   1.41575186
#> disp          0.02553724   0.02553821   0.02553913   0.02554002   0.02554087
#> drat          2.96862220   2.96860412   2.96858699   2.96857079   2.96855546
#> gear          5.06770622   5.06770250   5.06769915   5.06769615   5.06769350
#> hp            0.02865105   0.02865251   0.02865392   0.02865529   0.02865661
#> qsec          3.43332193   3.43340477   3.43348464   3.43356164   3.43363589
#> vs           -1.29962429  -1.29973913  -1.29985011  -1.29995734  -1.30006093
#> wt           -5.36942664  -5.36959934  -5.36976542  -5.36992509  -5.37007861
#>                   0.0058       0.0054
#> (Intercept) -68.83582981 -68.83677755
#> am            3.33722507   3.33720347
#> carb         -1.02720406  -1.02716367
#> cyl           1.41572332   1.41569601
#> disp          0.02554169   0.02554247
#> drat          2.96854097   2.96852727
#> gear          5.06769119   5.06768923
#> hp            0.02865788   0.02865912
#> qsec          3.43370747   3.43377647
#> vs           -1.30016099  -1.30025762
#> wt           -5.37022620  -5.37036807
#> 
#> $iter
#>   [1]   0   4   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   3   3   3   3   3   3   3   3  19  22  21  21  49  57
#>  [37]   1   1   1   1   1   1  32   1   1   1   1   1   1   1  82   1  70   1
#>  [55]   1   1 133   1   1   1   1 120   1   1   1   1   1   1   1   1   1   1
#>  [73]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.390408923 5.027113403 4.688302785 4.372326868 4.077646671 3.802826933
#>   [7] 3.546529125 3.307504932 3.084590169 2.876699116 2.682819223 2.502006187
#>  [13] 2.333379344 2.176117385 2.029454357 1.892675926 1.765115903 1.646152999
#>  [19] 1.535207796 1.431739929 1.335245449 1.245254375 1.161328398 1.083058751
#>  [25] 1.010064216 0.941989268 0.878502344 0.819294226 0.764076538 0.712580336
#>  [31] 0.664554807 0.619766037 0.577995880 0.539040892 0.502711340 0.468830278
#>  [37] 0.437232687 0.407764667 0.380282693 0.354652913 0.330750493 0.308459017
#>  [43] 0.287669912 0.268281922 0.250200619 0.233337934 0.217611739 0.202945436
#>  [49] 0.189267593 0.176511590 0.164615300 0.153520780 0.143173993 0.133524545
#>  [55] 0.124525437 0.116132840 0.108305875 0.101006422 0.094198927 0.087850234
#>  [61] 0.081929423 0.076407654 0.071258034 0.066455481 0.061976604 0.057799588
#>  [67] 0.053904089 0.050271134 0.046883028 0.043723269 0.040776467 0.038028269
#>  [73] 0.035465291 0.033075049 0.030845902 0.028766991 0.026828192 0.025020062
#>  [79] 0.023333793 0.021761174 0.020294544 0.018926759 0.017651159 0.016461530
#>  [85] 0.015352078 0.014317399 0.013352454 0.012452544 0.011613284 0.010830588
#>  [91] 0.010100642 0.009419893 0.008785023 0.008192942 0.007640765 0.007125803
#>  [97] 0.006645548 0.006197660 0.005779959 0.005390409
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 30
#> 
#> $loss
#>   [1] 803.09810 685.96070 587.95786 506.33595 438.71764 383.05166 337.56926
#>   [8] 300.74633 271.27062 248.01322 230.00384 216.40939 206.51537 199.70972
#>  [15] 195.46890 193.34572 192.91142 192.91142 192.91142 192.91142 192.91142
#>  [22] 192.91142 192.19422 186.58206 181.89669 178.00368 174.78756 172.14872
#>  [29] 170.00128 168.27130 164.53143 156.47144 150.81422 147.13557 133.80871
#>  [36] 115.40229 115.40229 115.40228 115.40228 115.40228 115.40228 115.40228
#>  [43]  99.35954  99.35953  99.35953  99.35952  99.35952  99.35952  99.35952
#>  [50]  99.35951  89.96753  89.96752  84.91880  84.91878  84.91877  84.91876
#>  [57]  77.93625  77.93624  77.93623  77.93622  77.93622  76.84025  76.84023
#>  [64]  76.84022  76.84020  76.84019  76.84018  76.84016  76.84015  76.84014
#>  [71]  76.84013  76.84013  76.84012  76.84011  76.84010  76.84010  76.84009
#>  [78]  76.84008  76.84008  76.84007  76.84007  76.84006  76.84006  76.84006
#>  [85]  76.84005  76.84005  76.84005  76.84005  76.84004  76.84004  76.84004
#>  [92]  76.84004  76.84003  76.84003  76.84003  76.84003  76.84003  76.84003
#>  [99]  76.84003  76.84003
#> 
#> $linear.predictors
#>      5.3904   5.0271   4.6883   4.3723   4.0776   3.8028   3.5465   3.3075
#> 1  21.70952 21.93092 22.13740 22.32996 22.50954 22.67702 22.83321 22.97888
#> 2  21.70952 21.79766 21.87987 21.95653 22.02802 22.09470 22.15688 22.21487
#> 3  21.70952 21.49196 21.28906 21.09983 20.92335 20.75877 20.60529 20.46214
#> 4  21.70952 21.63305 21.56174 21.49523 21.43320 21.37535 21.32140 21.27109
#> 5  21.70952 21.50241 21.30925 21.12912 20.96112 20.80445 20.65833 20.52206
#> 6  21.70952 21.50241 21.30925 21.12912 20.96112 20.80445 20.65833 20.52206
#> 7  21.70952 21.17319 20.67300 20.20652 19.77148 19.36576 18.98739 18.63451
#> 8  21.70952 20.46562 19.30555 18.22367 17.21470 16.27373 15.39618 14.57778
#> 9  21.70952 20.50690 19.38534 18.33936 17.36388 16.45414 15.60571 14.81447
#> 10 21.70952 22.15040 22.56157 22.94502 23.30263 23.63614 23.94718 24.23724
#> 11 21.70952 22.45611 23.15238 23.80172 24.40730 24.97207 25.49877 25.98997
#> 12 21.70952 22.34114 22.93019 23.47954 23.99187 24.46967 24.91526 25.33083
#> 13 21.70952 22.01192 22.29394 22.55695 22.80223 23.03098 23.24432 23.44327
#> 14 21.70952 21.46060 21.22846 21.01196 20.81006 20.62176 20.44615 20.28238
#> 15 21.70952 21.50502 21.31430 21.13644 20.97056 20.81586 20.67159 20.53704
#> 16 21.70952 21.29077 20.90023 20.53602 20.19635 19.87958 19.58415 19.30864
#> 17 21.70952 22.28889 22.82920 23.33310 23.80304 24.24130 24.65003 25.03121
#> 18 21.70952 22.18176 22.62216 23.03289 23.41593 23.77316 24.10631 24.41701
#> 19 21.70952 22.50941 23.25539 23.95109 24.59991 25.20500 25.76930 26.29557
#> 20 21.70952 21.85254 21.98591 22.11029 22.22629 22.33448 22.43537 22.52946
#> 21 21.70952 21.84731 21.97581 22.09565 22.20741 22.31164 22.40885 22.49950
#>      3.0846   2.8767   2.6828   2.5020   2.3334   2.1761   2.0295    1.8927
#> 1  23.11472 23.24142 23.35957 23.46976 23.57252 23.66836 23.75774 23.841095
#> 2  22.26895 22.31939 22.36643 22.41029 22.45121 22.48936 22.52494 22.558127
#> 3  20.32865 20.20415 20.08804 19.97976 19.87877 19.78459 19.69676 19.614848
#> 4  21.22417 21.18041 21.13960 21.10154 21.06605 21.03295 21.00208 20.973285
#> 5  20.39498 20.27646 20.16593 20.06285 19.96672 19.87706 19.79345 19.715473
#> 6  20.39498 20.27646 20.16593 20.06285 19.96672 19.87706 19.79345 19.715473
#> 7  18.30542 17.99851 17.71228 17.44535 17.19640 16.96423 16.74771 16.545788
#> 8  13.81453 13.10272 12.43889 11.81979 11.24242 10.70397 10.20180  9.733479
#> 9  14.07655 13.38837 12.74657 12.14802 11.58981 11.06923 10.58373 10.130948
#> 10 24.50776 24.76005 24.99534 25.21476 25.41940 25.61025 25.78823 25.954219
#> 11 26.44807 26.87529 27.27372 27.64530 27.99184 28.31502 28.61641 28.897498
#> 12 25.71838 26.07982 26.41689 26.73125 27.02442 27.29784 27.55282 27.790624
#> 13 23.62882 23.80187 23.96325 24.11375 24.25411 24.38501 24.50709 24.620939
#> 14 20.12964 19.98720 19.85436 19.73047 19.61493 19.50718 19.40669 19.312973
#> 15 20.41156 20.29454 20.18541 20.08362 19.98870 19.90018 19.81762 19.740629
#> 16 19.05169 18.81206 18.58859 18.38017 18.18580 18.00453 17.83548 17.677818
#> 17 25.38671 25.71824 26.02743 26.31578 26.58469 26.83548 27.06937 27.287499
#> 18 24.70677 24.97700 25.22902 25.46405 25.68324 25.88766 26.07830 26.256094
#> 19 26.78638 27.24410 27.67098 28.06909 28.44036 28.78662 29.10953 29.410685
#> 20 22.61721 22.69905 22.77537 22.84655 22.91293 22.97483 23.03257 23.086408
#> 21 22.58404 22.66289 22.73642 22.80500 22.86895 22.92860 22.98422 23.036096
#>       1.7651    1.6462    1.5352    1.4317    1.3352    1.2453    1.1613
#> 1  23.899522 23.899522 23.899522 23.899522 23.899522 23.899522 23.880881
#> 2  22.581388 22.581388 22.581388 22.581388 22.581388 22.581388 22.569504
#> 3  19.557433 19.557433 19.557433 19.557433 19.557433 19.557433 19.582259
#> 4  20.953104 20.953104 20.953104 20.953104 20.953104 20.953104 20.974905
#> 5  19.660816 19.660816 19.660816 19.660816 19.660816 19.660816 19.664338
#> 6  19.660816 19.660816 19.660816 19.660816 19.660816 19.660816 19.670982
#> 7  16.404249 16.404249 16.404249 16.404249 16.404249 16.404249 16.399180
#> 8   9.405216  9.405216  9.405216  9.405216  9.405216  9.405216  9.407750
#> 9   9.813579  9.813579  9.813579  9.813579  9.813579  9.813579  9.811512
#> 10 26.070566 26.070566 26.070566 26.070566 26.070566 26.070566 26.084339
#> 11 29.094521 29.094521 29.094521 29.094521 29.094521 29.094521 29.096498
#> 12 27.957307 27.957307 27.957307 27.957307 27.957307 27.957307 27.975045
#> 13 24.700741 24.700741 24.700741 24.700741 24.700741 24.700741 24.721071
#> 14 19.247284 19.247284 19.247284 19.247284 19.247284 19.247284 19.235146
#> 15 19.686662 19.686662 19.686662 19.686662 19.686662 19.686662 19.679100
#> 16 17.567309 17.567309 17.567309 17.567309 17.567309 17.567309 17.557873
#> 17 27.440392 27.440392 27.440392 27.440392 27.440392 27.440392 27.447275
#> 18 26.380715 26.380715 26.380715 26.380715 26.380715 26.380715 26.363685
#> 19 29.621775 29.621775 29.621775 29.621775 29.621775 29.621775 29.605591
#> 20 23.124149 23.124149 23.124149 23.124149 23.124149 23.124149 23.095205
#> 21 23.072457 23.072457 23.072457 23.072457 23.072457 23.072457 23.077862
#>       1.0831    1.0101    0.9420    0.8785    0.8193    0.7641    0.7126
#> 1  23.726156 23.581874 23.447317 23.321828 23.204797 23.095654 22.993866
#> 2  22.470865 22.378883 22.293101 22.213100 22.138492 22.068911 22.004020
#> 3  19.788320 19.980473 20.159675 20.326800 20.482661 20.628018 20.763578
#> 4  21.155858 21.324598 21.481965 21.628726 21.765595 21.893241 22.012283
#> 5  19.693572 19.720832 19.746255 19.769965 19.792076 19.812698 19.831929
#> 6  19.755361 19.834045 19.907425 19.975860 20.039683 20.099205 20.154714
#> 7  16.357101 16.317862 16.281267 16.247139 16.215311 16.185628 16.157946
#> 8   9.428786  9.448402  9.466695  9.483756  9.499667  9.514506  9.528345
#> 9   9.794353  9.778353  9.763432  9.749515  9.736537  9.724434  9.713146
#> 10 26.198657 26.305259 26.404677 26.497394 26.583862 26.664503 26.739708
#> 11 29.112912 29.128217 29.142491 29.155804 29.168218 29.179797 29.190594
#> 12 28.122276 28.259570 28.387610 28.507021 28.618384 28.722242 28.819100
#> 13 24.889818 25.047174 25.193926 25.330787 25.458424 25.577459 25.688471
#> 14 19.134397 19.040449 18.952833 18.871121 18.794917 18.723849 18.657570
#> 15 19.616334 19.557804 19.503219 19.452313 19.404837 19.360562 19.319270
#> 16 17.479552 17.406518 17.338405 17.274884 17.215643 17.160395 17.108871
#> 17 27.504407 27.557683 27.607369 27.653706 27.696919 27.737221 27.774806
#> 18 26.222328 26.090513 25.967581 25.852935 25.746015 25.646301 25.553308
#> 19 29.471265 29.346006 29.229188 29.120244 29.018642 28.923888 28.835519
#> 20 22.854963 22.630936 22.422008 22.227161 22.045446 21.875978 21.717932
#> 21 23.122719 23.164549 23.203560 23.239941 23.273870 23.305513 23.335023
#>       0.6646    0.6198    0.5780    0.5390   0.5027   0.4688   0.4372   0.4078
#> 1  22.915563 22.884031 22.854536 22.827088 22.20818 21.07276 21.07242 21.07215
#> 2  21.973170 22.018988 22.061514 22.101312 21.44657 20.65839 20.65813 20.65793
#> 3  20.756966 20.415603 20.098222 19.801572 18.61827 18.86957 18.86969 18.86978
#> 4  22.124387 22.232223 22.332735 22.426511 21.95534 23.22214 23.22246 23.22270
#> 5  19.911464 20.141222 20.355012 20.554720 19.83799 19.80670 19.80665 19.80660
#> 6  20.265227 20.516834 20.751008 20.969721 19.83799 19.80670 19.80665 19.80660
#> 7  16.099993 15.964695 15.838775 15.721166 15.74138 15.13491 15.13477 15.13467
#> 8   9.579254  9.722729  9.856235  9.980946 10.28203 10.81545 10.81557 10.81567
#> 9   9.774684 10.013935 10.236509 10.444456 11.42556 11.70793 11.70801 11.70808
#> 10 26.815247 26.899679 26.978348 27.051764 27.59025 28.01321 28.01337 28.01350
#> 11 29.311379 29.703549 30.068428 30.409297 31.55246 29.90106 29.90075 29.90050
#> 12 28.912837 29.009306 29.099208 29.183096 29.23284 29.07144 29.07147 29.07149
#> 13 25.749987 25.701815 25.657170 25.615344 25.29929 23.60932 23.60904 23.60881
#> 14 18.482362 18.032409 17.613685 17.222571 17.29994 16.88542 16.88534 16.88527
#> 15 19.222862 18.986606 18.766737 18.561372 19.09275 18.31921 18.31907 18.31896
#> 16 17.018098 16.825363 16.645969 16.478427 17.59198 17.42040 17.42041 17.42042
#> 17 27.801649 27.806132 27.810363 27.814275 28.38173 28.44383 28.44390 28.44396
#> 18 25.537812 25.702736 25.856030 25.999341 28.01143 30.09765 30.09812 30.09849
#> 19 28.678519 28.343407 28.031493 27.740187 27.27974 28.92387 28.92416 28.92439
#> 20 21.555898 21.367100 21.191202 21.027040 18.37705 18.83631 18.83617 18.83605
#> 21 23.412644 23.611638 23.796821 23.969794 24.83721 25.28372 25.28387 25.28398
#>      0.3803   0.3547   0.3308   0.3085   0.2877   0.2683   0.2502   0.2333
#> 1  21.07194 21.07177 21.07163 21.07153 21.47361 21.47338 21.47316 21.47295
#> 2  20.65777 20.65765 20.65755 20.65747 21.57494 21.57483 21.57470 21.57455
#> 3  18.86985 18.86991 18.86996 18.86999 20.22930 20.22925 20.22920 20.22916
#> 4  23.22290 23.22305 23.22317 23.22327 23.01149 23.01167 23.01185 23.01202
#> 5  19.80657 19.80654 19.80651 19.80650 18.80450 18.80439 18.80432 18.80429
#> 6  19.80657 19.80654 19.80651 19.80650 19.27960 19.27949 19.27942 19.27937
#> 7  15.13459 15.13452 15.13447 15.13443 14.79848 14.79831 14.79818 14.79807
#> 8  10.81575 10.81581 10.81586 10.81590 11.71421 11.71445 11.71462 11.71474
#> 9  11.70813 11.70817 11.70820 11.70823 11.92319 11.92342 11.92359 11.92371
#> 10 28.01360 28.01368 28.01375 28.01380 29.16678 29.16703 29.16720 29.16733
#> 11 29.90030 29.90015 29.90002 29.89993 29.27792 29.27777 29.27763 29.27751
#> 12 29.07151 29.07152 29.07153 29.07154 30.25243 30.25252 30.25257 30.25259
#> 13 23.60863 23.60848 23.60837 23.60828 23.11328 23.11287 23.11255 23.11230
#> 14 16.88522 16.88518 16.88515 16.88512 16.02355 16.02331 16.02313 16.02302
#> 15 18.31888 18.31881 18.31876 18.31871 17.18979 17.18954 17.18937 17.18926
#> 16 17.42043 17.42043 17.42044 17.42044 16.31404 16.31398 16.31396 16.31396
#> 17 28.44400 28.44404 28.44407 28.44409 29.07095 29.07107 29.07116 29.07121
#> 18 30.09878 30.09901 30.09919 30.09934 28.58963 28.59013 28.59058 28.59097
#> 19 28.92457 28.92471 28.92483 28.92492 28.38473 28.38487 28.38501 28.38515
#> 20 18.83596 18.83589 18.83583 18.83579 19.42566 19.42551 19.42536 19.42523
#> 21 25.28407 25.28414 25.28419 25.28424 26.28192 26.28222 26.28243 26.28259
#>      0.2176   0.2029   0.1893   0.1765   0.1646   0.1535   0.1432   0.1335
#> 1  21.47277 21.47261 21.47247 21.47235 21.10392 21.10384 20.29429 20.29401
#> 2  21.57442 21.57429 21.57417 21.57407 21.62967 21.62970 20.86955 20.86937
#> 3  20.22913 20.22910 20.22908 20.22907 20.54708 20.54728 20.25833 20.25840
#> 4  23.01217 23.01231 23.01243 23.01253 22.35464 22.35450 21.26361 21.26324
#> 5  18.80428 18.80427 18.80428 18.80428 19.23747 19.23746 19.16203 19.16203
#> 6  19.27934 19.27933 19.27932 19.27931 20.27715 20.27726 20.53193 20.53211
#> 7  14.79799 14.79793 14.79788 14.79784 15.07284 15.07289 15.31609 15.31619
#> 8  11.71483 11.71488 11.71492 11.71495 11.46319 11.46317 11.55275 11.55275
#> 9  11.92380 11.92387 11.92392 11.92396 11.51539 11.51529 12.08477 12.08481
#> 10 29.16742 29.16748 29.16752 29.16755 29.66224 29.66239 29.61719 29.61730
#> 11 29.27741 29.27732 29.27725 29.27718 29.48974 29.48965 29.47773 29.47767
#> 12 30.25259 30.25258 30.25257 30.25256 31.41357 31.41383 32.02694 32.02731
#> 13 23.11212 23.11197 23.11186 23.11177 21.16085 21.16040 21.77446 21.77419
#> 14 16.02293 16.02288 16.02285 16.02282 16.08624 16.08629 15.54079 15.54070
#> 15 17.18918 17.18912 17.18909 17.18906 18.00978 18.00992 18.02018 18.02032
#> 16 16.31398 16.31401 16.31404 16.31406 16.67681 16.67686 16.89590 16.89602
#> 17 29.07125 29.07128 29.07130 29.07131 29.13659 29.13664 29.04667 29.04669
#> 18 28.59131 28.59160 28.59184 28.59204 28.39322 28.39313 28.06563 28.06553
#> 19 28.38528 28.38540 28.38550 28.38559 28.08951 28.08952 28.82142 28.82161
#> 20 19.42510 19.42499 19.42490 19.42482 19.14914 19.14916 19.65522 19.65532
#> 21 26.28269 26.28277 26.28282 26.28286 25.43097 25.43083 25.62451 25.62442
#>      0.1245   0.1161   0.1083   0.1010   0.0942   0.0879   0.0819   0.0764
#> 1  20.29375 20.29351 20.46010 20.46019 20.46029 20.46038 20.46046 20.42335
#> 2  20.86921 20.86906 20.83422 20.83421 20.83421 20.83422 20.83422 20.97675
#> 3  20.25847 20.25853 20.39034 20.39022 20.39010 20.38999 20.38988 20.33182
#> 4  21.26290 21.26259 21.65545 21.65546 21.65548 21.65550 21.65552 21.76506
#> 5  19.16202 19.16202 18.68340 18.68334 18.68329 18.68325 18.68321 18.33010
#> 6  20.53228 20.53244 20.42902 20.42886 20.42871 20.42856 20.42843 20.38705
#> 7  15.31629 15.31637 13.68762 13.68763 13.68764 13.68766 13.68768 13.82051
#> 8  11.55274 11.55274 12.43366 12.43368 12.43369 12.43370 12.43370 12.45816
#> 9  12.08485 12.08489 12.14052 12.14063 12.14073 12.14082 12.14090 12.11241
#> 10 29.61741 29.61750 29.16701 29.16688 29.16675 29.16663 29.16652 29.16266
#> 11 29.47760 29.47754 30.26688 30.26690 30.26692 30.26694 30.26695 30.05802
#> 12 32.02765 32.02797 32.40121 32.40099 32.40078 32.40058 32.40038 32.78631
#> 13 21.77394 21.77371 21.71588 21.71639 21.71688 21.71734 21.71776 21.99916
#> 14 15.54063 15.54055 15.44914 15.44910 15.44906 15.44902 15.44899 15.27844
#> 15 18.02046 18.02058 17.88242 17.88227 17.88213 17.88200 17.88187 18.00977
#> 16 16.89614 16.89624 17.93175 17.93166 17.93156 17.93147 17.93137 17.90730
#> 17 29.04670 29.04671 28.82066 28.82063 28.82060 28.82057 28.82054 28.63620
#> 18 28.06544 28.06535 27.67844 27.67837 27.67831 27.67826 27.67822 28.13165
#> 19 28.82179 28.82195 29.29300 29.29303 29.29304 29.29305 29.29306 28.91035
#> 20 19.65541 19.65549 19.76039 19.76050 19.76061 19.76070 19.76079 19.97838
#> 21 25.62433 25.62425 24.81888 24.81905 24.81922 24.81939 24.81954 24.43654
#>      0.0713   0.0665   0.0620   0.0578   0.0539   0.0503   0.0469   0.0437
#> 1  20.42332 20.42330 20.42328 20.42326 20.42324 20.42323 20.42321 20.42319
#> 2  20.97673 20.97671 20.97669 20.97667 20.97666 20.97664 20.97662 20.97661
#> 3  20.33184 20.33186 20.33187 20.33188 20.33189 20.33190 20.33191 20.33192
#> 4  21.76501 21.76496 21.76491 21.76487 21.76483 21.76479 21.76475 21.76471
#> 5  18.32990 18.32972 18.32953 18.32936 18.32918 18.32902 18.32886 18.32870
#> 6  20.38701 20.38698 20.38694 20.38690 20.38687 20.38683 20.38680 20.38676
#> 7  13.82035 13.82019 13.82003 13.81989 13.81975 13.81961 13.81948 13.81935
#> 8  12.45830 12.45844 12.45857 12.45870 12.45882 12.45894 12.45905 12.45916
#> 9  12.11246 12.11251 12.11256 12.11260 12.11265 12.11269 12.11274 12.11278
#> 10 29.16257 29.16248 29.16239 29.16231 29.16223 29.16215 29.16207 29.16200
#> 11 30.05805 30.05807 30.05810 30.05812 30.05815 30.05817 30.05819 30.05821
#> 12 32.78650 32.78668 32.78685 32.78702 32.78718 32.78733 32.78748 32.78762
#> 13 21.99939 21.99961 21.99983 22.00004 22.00025 22.00046 22.00066 22.00085
#> 14 15.27834 15.27825 15.27815 15.27806 15.27797 15.27789 15.27780 15.27772
#> 15 18.00976 18.00974 18.00972 18.00970 18.00968 18.00966 18.00965 18.00963
#> 16 17.90743 17.90756 17.90768 17.90779 17.90790 17.90800 17.90810 17.90820
#> 17 28.63611 28.63602 28.63594 28.63586 28.63578 28.63570 28.63563 28.63556
#> 18 28.13157 28.13149 28.13141 28.13134 28.13126 28.13120 28.13113 28.13107
#> 19 28.91042 28.91048 28.91053 28.91059 28.91064 28.91069 28.91073 28.91078
#> 20 19.97858 19.97877 19.97896 19.97914 19.97931 19.97948 19.97965 19.97981
#> 21 24.43637 24.43621 24.43606 24.43591 24.43576 24.43563 24.43549 24.43537
#>      0.0408   0.0380   0.0355   0.0331   0.0308   0.0288   0.0268   0.0250
#> 1  20.42318 20.42316 20.42315 20.42313 20.42312 20.42310 20.42309 20.42308
#> 2  20.97659 20.97658 20.97657 20.97655 20.97654 20.97653 20.97652 20.97650
#> 3  20.33192 20.33193 20.33194 20.33194 20.33194 20.33195 20.33195 20.33195
#> 4  21.76467 21.76464 21.76461 21.76457 21.76454 21.76451 21.76449 21.76446
#> 5  18.32855 18.32841 18.32827 18.32813 18.32800 18.32787 18.32775 18.32764
#> 6  20.38673 20.38670 20.38667 20.38664 20.38661 20.38658 20.38656 20.38653
#> 7  13.81923 13.81912 13.81901 13.81890 13.81880 13.81871 13.81861 13.81852
#> 8  12.45927 12.45937 12.45947 12.45956 12.45965 12.45974 12.45982 12.45990
#> 9  12.11282 12.11286 12.11290 12.11294 12.11297 12.11301 12.11304 12.11308
#> 10 29.16193 29.16186 29.16179 29.16173 29.16166 29.16160 29.16154 29.16149
#> 11 30.05822 30.05824 30.05826 30.05827 30.05829 30.05830 30.05831 30.05833
#> 12 32.78776 32.78789 32.78802 32.78814 32.78826 32.78837 32.78848 32.78859
#> 13 22.00105 22.00123 22.00141 22.00159 22.00176 22.00193 22.00209 22.00225
#> 14 15.27764 15.27757 15.27749 15.27742 15.27735 15.27728 15.27722 15.27715
#> 15 18.00961 18.00960 18.00958 18.00956 18.00955 18.00953 18.00952 18.00950
#> 16 17.90829 17.90838 17.90846 17.90854 17.90862 17.90869 17.90876 17.90883
#> 17 28.63549 28.63542 28.63535 28.63529 28.63523 28.63517 28.63511 28.63506
#> 18 28.13101 28.13096 28.13090 28.13085 28.13080 28.13076 28.13071 28.13067
#> 19 28.91082 28.91086 28.91090 28.91094 28.91097 28.91100 28.91104 28.91107
#> 20 19.97996 19.98011 19.98025 19.98039 19.98053 19.98066 19.98079 19.98091
#> 21 24.43524 24.43512 24.43501 24.43490 24.43480 24.43469 24.43460 24.43450
#>      0.0233   0.0218   0.0203   0.0189   0.0177   0.0165   0.0154   0.0143
#> 1  20.42307 20.42305 20.42304 20.42303 20.42302 20.42301 20.42300 20.42299
#> 2  20.97649 20.97648 20.97647 20.97646 20.97645 20.97644 20.97643 20.97642
#> 3  20.33195 20.33195 20.33195 20.33195 20.33195 20.33195 20.33195 20.33195
#> 4  21.76443 21.76441 21.76438 21.76436 21.76433 21.76431 21.76429 21.76427
#> 5  18.32752 18.32741 18.32731 18.32721 18.32711 18.32701 18.32692 18.32683
#> 6  20.38651 20.38648 20.38646 20.38644 20.38642 20.38640 20.38638 20.38636
#> 7  13.81844 13.81836 13.81828 13.81821 13.81813 13.81807 13.81800 13.81794
#> 8  12.45998 12.46005 12.46013 12.46019 12.46026 12.46032 12.46038 12.46044
#> 9  12.11311 12.11314 12.11317 12.11320 12.11323 12.11326 12.11329 12.11331
#> 10 29.16143 29.16138 29.16133 29.16128 29.16123 29.16118 29.16114 29.16110
#> 11 30.05834 30.05835 30.05836 30.05837 30.05838 30.05838 30.05839 30.05840
#> 12 32.78869 32.78879 32.78888 32.78898 32.78906 32.78915 32.78923 32.78931
#> 13 22.00240 22.00255 22.00270 22.00284 22.00297 22.00311 22.00323 22.00336
#> 14 15.27709 15.27703 15.27697 15.27691 15.27686 15.27681 15.27675 15.27670
#> 15 18.00949 18.00947 18.00946 18.00945 18.00944 18.00942 18.00941 18.00940
#> 16 17.90889 17.90895 17.90901 17.90907 17.90912 17.90917 17.90922 17.90927
#> 17 28.63500 28.63495 28.63490 28.63485 28.63480 28.63476 28.63471 28.63467
#> 18 28.13063 28.13059 28.13055 28.13052 28.13049 28.13045 28.13042 28.13040
#> 19 28.91109 28.91112 28.91115 28.91117 28.91119 28.91122 28.91124 28.91126
#> 20 19.98103 19.98114 19.98126 19.98136 19.98147 19.98157 19.98166 19.98175
#> 21 24.43441 24.43433 24.43424 24.43416 24.43408 24.43401 24.43394 24.43387
#>      0.0134   0.0125   0.0116   0.0108   0.0101   0.0094   0.0088   0.0082
#> 1  20.42298 20.42296 20.42295 20.42294 20.42293 20.42293 20.42292 20.42291
#> 2  20.97642 20.97641 20.97640 20.97639 20.97638 20.97638 20.97637 20.97636
#> 3  20.33195 20.33195 20.33195 20.33195 20.33195 20.33194 20.33194 20.33194
#> 4  21.76425 21.76423 21.76421 21.76419 21.76418 21.76416 21.76414 21.76413
#> 5  18.32675 18.32667 18.32659 18.32651 18.32644 18.32637 18.32631 18.32624
#> 6  20.38634 20.38632 20.38631 20.38629 20.38627 20.38626 20.38625 20.38623
#> 7  13.81788 13.81782 13.81777 13.81772 13.81767 13.81762 13.81757 13.81753
#> 8  12.46050 12.46055 12.46060 12.46065 12.46070 12.46075 12.46079 12.46083
#> 9  12.11334 12.11336 12.11339 12.11341 12.11343 12.11345 12.11347 12.11349
#> 10 29.16106 29.16102 29.16098 29.16094 29.16091 29.16087 29.16084 29.16081
#> 11 30.05840 30.05841 30.05842 30.05842 30.05843 30.05843 30.05843 30.05844
#> 12 32.78939 32.78946 32.78953 32.78960 32.78967 32.78973 32.78979 32.78985
#> 13 22.00348 22.00359 22.00370 22.00381 22.00391 22.00401 22.00411 22.00421
#> 14 15.27665 15.27661 15.27656 15.27652 15.27648 15.27643 15.27640 15.27636
#> 15 18.00939 18.00938 18.00937 18.00936 18.00935 18.00934 18.00934 18.00933
#> 16 17.90931 17.90936 17.90940 17.90944 17.90948 17.90951 17.90955 17.90958
#> 17 28.63463 28.63459 28.63455 28.63451 28.63448 28.63444 28.63441 28.63438
#> 18 28.13037 28.13034 28.13032 28.13030 28.13027 28.13025 28.13023 28.13022
#> 19 28.91127 28.91129 28.91131 28.91132 28.91134 28.91135 28.91137 28.91138
#> 20 19.98184 19.98193 19.98201 19.98209 19.98217 19.98225 19.98232 19.98239
#> 21 24.43380 24.43374 24.43367 24.43362 24.43356 24.43350 24.43345 24.43340
#>      0.0076   0.0071   0.0066   0.0062   0.0058   0.0054
#> 1  20.42290 20.42289 20.42288 20.42287 20.42286 20.42285
#> 2  20.97636 20.97635 20.97634 20.97634 20.97633 20.97633
#> 3  20.33194 20.33194 20.33194 20.33193 20.33193 20.33193
#> 4  21.76411 21.76410 21.76408 21.76407 21.76406 21.76404
#> 5  18.32618 18.32612 18.32606 18.32601 18.32595 18.32590
#> 6  20.38622 20.38621 20.38620 20.38619 20.38618 20.38617
#> 7  13.81749 13.81745 13.81741 13.81738 13.81734 13.81731
#> 8  12.46087 12.46091 12.46095 12.46098 12.46102 12.46105
#> 9  12.11351 12.11353 12.11355 12.11356 12.11358 12.11360
#> 10 29.16078 29.16075 29.16072 29.16070 29.16067 29.16065
#> 11 30.05844 30.05845 30.05845 30.05845 30.05845 30.05846
#> 12 32.78991 32.78997 32.79002 32.79007 32.79012 32.79017
#> 13 22.00430 22.00438 22.00447 22.00455 22.00462 22.00470
#> 14 15.27632 15.27628 15.27625 15.27622 15.27618 15.27615
#> 15 18.00932 18.00931 18.00931 18.00930 18.00929 18.00929
#> 16 17.90961 17.90964 17.90967 17.90970 17.90972 17.90975
#> 17 28.63435 28.63432 28.63429 28.63426 28.63423 28.63421
#> 18 28.13020 28.13018 28.13016 28.13015 28.13014 28.13012
#> 19 28.91139 28.91140 28.91141 28.91142 28.91143 28.91144
#> 20 19.98245 19.98252 19.98258 19.98264 19.98270 19.98275
#> 21 24.43335 24.43330 24.43326 24.43321 24.43317 24.43313
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 18.1 24.4 19.2 17.8 16.4 10.4 14.7 32.4 30.4 33.9 21.5 15.5 15.2
#> [16] 19.2 27.3 26.0 30.4 19.7 21.4
#> 
#> $X
#>               am       carb        cyl       disp        drat       gear
#>  [1,]  1.0488088  1.0463702  0.1714986 -0.3077480  0.39152050  0.3483665
#>  [2,]  1.0488088  1.0463702  0.1714986 -0.3077480  0.39152050  0.3483665
#>  [3,] -0.9534626 -1.1510072  0.1714986  0.2373098 -1.65498620 -1.1147728
#>  [4,] -0.9534626 -0.4185481 -1.0289915 -0.4192752  0.01453242  0.3483665
#>  [5,] -0.9534626  1.0463702  0.1714986 -0.2440182  0.42742412  0.3483665
#>  [6,] -0.9534626  1.0463702  0.1714986 -0.2440182  0.42742412  0.3483665
#>  [7,] -0.9534626  0.3139110  1.3719887  0.6632934 -1.09847999 -1.1147728
#>  [8,] -0.9534626  1.0463702  1.3719887  2.2079034 -1.22414269 -1.1147728
#>  [9,] -0.9534626  1.0463702  1.3719887  2.0401933 -0.81125098 -1.1147728
#> [10,]  1.0488088 -1.1510072 -1.0289915 -0.9894896  0.71465313  0.3483665
#> [11,]  1.0488088 -0.4185481 -1.0289915 -1.0146461  2.24055725  0.3483665
#> [12,]  1.0488088 -1.1510072 -1.0289915 -1.0532194  0.96597852  0.3483665
#> [13,] -0.9534626 -1.1510072 -1.0289915 -0.6423297  0.03248423 -1.1147728
#> [14,] -0.9534626 -0.4185481  1.3719887  1.0171617 -1.65498620 -1.1147728
#> [15,] -0.9534626 -0.4185481  1.3719887  0.8997647 -0.95486549 -1.1147728
#> [16,] -0.9534626 -0.4185481  1.3719887  1.7047731 -1.08052818 -1.1147728
#> [17,]  1.0488088 -1.1510072 -1.0289915 -0.9869739  0.71465313  0.3483665
#> [18,]  1.0488088 -0.4185481 -1.0289915 -0.6406526  1.34296659  1.8115058
#> [19,]  1.0488088 -0.4185481 -1.0289915 -0.8519673  0.15814693  1.8115058
#> [20,]  1.0488088  2.5112884  0.1714986 -0.4335306 -0.11113027  1.8115058
#> [21,]  1.0488088 -0.4185481 -1.0289915 -0.6347827  0.76850857  0.3483665
#>                hp       qsec         vs         wt
#>  [1,] -0.24761336 -1.2072807 -1.0488088 -0.4062768
#>  [2,] -0.24761336 -0.7847325 -1.0488088 -0.1617435
#>  [3,] -0.34877835  1.6298290  0.9534626  0.3992444
#>  [4,] -1.21879726  1.4638279  0.9534626  0.1403269
#>  [5,]  0.01541562  0.1810921  0.9534626  0.3800654
#>  [6,]  0.01541562  0.6338224  0.9534626  0.3800654
#>  [7,]  1.16869650 -0.4980033 -1.0488088  0.9842063
#>  [8,]  1.87685143 -0.1810921 -1.0488088  2.2826298
#>  [9,]  2.18034640 -0.4829123 -1.0488088  2.2068724
#> [10,] -1.13786526  1.0639161  0.9534626 -0.8090374
#> [11,] -1.42112724  0.3470932  0.9534626 -1.3700254
#> [12,] -1.15809826  1.3883728  0.9534626 -1.1590555
#> [13,] -0.51064233  1.4713734  0.9534626 -0.5549146
#> [14,]  0.56170656 -0.8979150 -1.0488088  0.4567817
#> [15,]  0.56170656 -0.5734583 -1.0488088  0.3752706
#> [16,]  1.06753151 -0.7620960 -1.0488088  0.7684417
#> [17,] -1.13786526  0.6338224  0.9534626 -1.0631602
#> [18,] -0.63204032 -1.0261886 -1.0488088 -0.8665746
#> [19,] -0.18691436 -0.8752785  0.9534626 -1.4678387
#> [20,]  1.06753151 -1.9316492 -1.0488088 -0.2624337
#> [21,] -0.26784635  0.4074572  0.9534626 -0.2528442
#> attr(,"center")
#>  [1]   0.4761905   2.5714286   5.7142857 196.7000000   3.6819048   3.7619048
#>  [7] 122.2380952  18.0600000   0.5238095   3.0436667
#> attr(,"scale")
#>  [1]   0.4994328   1.3652639   1.6659863 119.2534037   0.5570468   0.6834619
#>  [7]  49.4242130   1.3252924   0.4994328   1.0428031
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 19
#> 
#> $lambda.min
#> [1] 1.535208
#> 
#> $null.dev
#> [1] 38.24277
#> 
#> $Bias
#> [1] 8.803535
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 9.362388