Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1]  46.25168  43.12689  38.55533  33.89494  30.11460  27.08129  24.68053
#>   [8]  22.81394  21.39701  20.35726  19.63258  19.16981  18.92356  18.85509
#>  [15]  18.93143  19.11273  18.89263  18.53626  18.48261  18.33257  18.18242
#>  [22]  18.06417  17.97282  17.77296  16.85823  16.63086  14.58809  12.12319
#>  [29]  11.99819  11.94595  12.25243  12.19621  13.74797  13.62856  13.52784
#>  [36]  13.41406  13.34916  13.21180  15.29267  14.99672  16.93966  17.39609
#>  [43]  17.31307  17.23851  17.20762  16.20485  26.58206  26.66904  31.39727
#>  [50]  31.13043  38.93590  38.91345  38.69286 151.85579 152.74224 162.66443
#>  [57] 162.92775 163.11619 163.91441 163.92613 164.18408 164.19552 164.20675
#>  [64] 164.28135 164.75144 164.85780 164.74052 168.75795 168.71503 168.70057
#>  [71] 169.40598 178.72871 178.72468 178.69918 178.69512 178.69103 178.68842
#>  [78] 178.70066 178.69661 178.69233 178.74333 178.73960 178.71719 178.78101
#>  [85] 179.13726 179.13330 179.04071 179.03663 179.03257 179.02852 179.02449
#>  [92] 179.02047 179.01647 179.01249 179.00852 179.00457 179.00067 178.99687
#>  [99] 178.99065 178.98669
#> 
#> $cvse
#>   [1] 11.171264 11.042147 10.294527  9.333311  8.467622  7.683961  6.974406
#>   [8]  6.335626  5.768071  5.275204  4.862596  4.536618  4.302602  4.162667
#>  [15]  4.113974  4.139878  4.100323  4.022678  4.000620  3.968007  3.932598
#>  [22]  3.901835  3.875271  3.835931  3.745432  3.711268  3.591485  3.209129
#>  [29]  3.192608  3.176061  3.136508  3.117283  3.315532  3.291523  3.276759
#>  [36]  3.280952  3.280520  3.285969  4.697423  4.677071  6.396334  6.375872
#>  [43]  6.369889  6.365108  6.363557  6.221753 12.119934 12.109042 14.260359
#>  [50] 14.275567 17.838082 17.835910 17.844202 68.828922 68.750382 69.742428
#>  [57] 69.721656 69.710235 69.638253 69.643266 69.622165 69.627045 69.631866
#>  [64] 69.632062 69.583928 69.579250 69.597221 71.416711 71.426703 71.433614
#>  [71] 71.855008 76.236429 76.233415 76.232657 76.229663 76.226681 76.224072
#>  [78] 76.219504 76.216546 76.213621 76.205653 76.202695 76.201763 76.190826
#>  [85] 76.240044 76.237125 76.242714 76.239840 76.236982 76.234139 76.231312
#>  [92] 76.228500 76.225703 76.222922 76.220156 76.217406 76.214668 76.211949
#>  [99] 76.209498 76.206817
#> 
#> $fold
#>  [1]  3  1  5  1  4  6  2  8 10 10  4  1  7  8  9  9  5  6  7  3  2
#> 
#> $lambda
#>   [1] 5.651843037 5.270927733 4.915684845 4.584384138 4.275411991 3.987263532
#>   [7] 3.718535315 3.467918480 3.234192381 3.016218639 2.812935598 2.623353153
#>  [13] 2.446547931 2.281658788 2.127882623 1.984470456 1.850723789 1.725991199
#>  [19] 1.609665169 1.501179124 1.400004676 1.305649046 1.217652670 1.135586956
#>  [25] 1.059052195 0.987675622 0.921109591 0.859029887 0.801134148 0.747140388
#>  [31] 0.696785626 0.649824607 0.606028603 0.565184304 0.527092773 0.491568484
#>  [37] 0.458438414 0.427541199 0.398726353 0.371853531 0.346791848 0.323419238
#>  [43] 0.301621864 0.281293560 0.262335315 0.244654793 0.228165879 0.212788262
#>  [49] 0.198447046 0.185072379 0.172599120 0.160966517 0.150117912 0.140000468
#>  [55] 0.130564905 0.121765267 0.113558696 0.105905220 0.098767562 0.092110959
#>  [61] 0.085902989 0.080113415 0.074714039 0.069678563 0.064982461 0.060602860
#>  [67] 0.056518430 0.052709277 0.049156848 0.045843841 0.042754120 0.039872635
#>  [73] 0.037185353 0.034679185 0.032341924 0.030162186 0.028129356 0.026233532
#>  [79] 0.024465479 0.022816588 0.021278826 0.019844705 0.018507238 0.017259912
#>  [85] 0.016096652 0.015011791 0.014000047 0.013056490 0.012176527 0.011355870
#>  [91] 0.010590522 0.009876756 0.009211096 0.008590299 0.008011341 0.007471404
#>  [97] 0.006967856 0.006498246 0.006060286 0.005651843
#> 
#> $fit
#> $beta
#>              5.6518     5.2709    4.9157    4.5844    4.2754    3.9873
#> (Intercept) 21.5619 23.2170910 24.760723 26.200320 27.542893 28.794980
#> am           0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> cyl          0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> disp         0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.0000  0.0000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.0000 -0.5355599 -1.035025 -1.500828 -1.935237 -2.340368
#>                3.7185    3.4679    3.2342   3.0162    2.8129    2.6234
#> (Intercept) 29.962681 31.051683 32.067290 33.01445 33.897772 34.721562
#> am           0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> disp         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.00000  0.000000  0.000000
#> wt          -2.718195 -3.070558 -3.399173 -3.70564 -3.991452 -4.258001
#>                2.4465    2.2817    2.1279    1.9845    1.8507       1.7260
#> (Intercept) 35.489832 36.206322 36.874524 37.497691 37.934491 37.903905388
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.000000  0.000000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.001631224
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000000
#> wt          -4.506586 -4.738417 -4.954624 -5.156259 -5.297592 -5.297144242
#>                 1.6097     1.5012     1.4000     1.3056     1.2177     1.1356
#> (Intercept) 35.5355818 33.3281808 31.2695516 29.3496670 27.5591763 25.8893586
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.1279411  0.2456685  0.3554614  0.4578547  0.5533469  0.6424034
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -5.2624787 -5.2301686 -5.2000361 -5.1719344 -5.1457267 -5.1212853
#>                 1.0591     0.9877     0.9211     0.8590    0.8011    0.7471
#> (Intercept) 24.3320811 22.8797588 21.5253182 20.2621624 19.084139 17.985511
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> qsec         0.7254577  0.8029144  0.8751509  0.9425188  1.005346  1.063940
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.000000
#> wt          -5.0984911 -5.0772332 -5.0574081 -5.0389190 -5.021676 -5.005595
#>                0.6968    0.6498    0.6060    0.5652   0.5271    0.4916
#> (Intercept) 16.960926 16.005395 15.114263 14.984735 14.98454 14.984531
#> am           0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> cyl          0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> disp         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> qsec         1.118584  1.169545  1.217072  1.223980  1.22399  1.223991
#> vs           0.000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> wt          -4.990598 -4.976612 -4.963568 -4.961672 -4.96167 -4.961669
#>                 0.4584    0.4275     0.3987     0.3719     0.3468     0.3234
#> (Intercept) 10.4795179  1.633555 -0.3477571 -0.3527678 -0.3560164 -0.3581225
#> am           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.6567274  1.945909  2.2346437  2.2353097  2.2357415  2.2360214
#> gear         0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         1.2927812  1.427950  1.4582287  1.4583220  1.4583825  1.4584218
#> vs           0.0000000  0.000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -4.6843450 -4.139911 -4.0179734 -4.0176857 -4.0174992 -4.0173783
#>                 0.3016     0.2813     0.2623     0.2447     0.2282     0.2128
#> (Intercept) -0.3594879 -0.3603731 -0.3609471 -0.3613192 -0.3615604 -0.3617168
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         2.2362029  2.2363206  2.2363969  2.2364463  2.2364784  2.2364992
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         1.4584472  1.4584637  1.4584744  1.4584813  1.4584858  1.4584887
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -4.0172999 -4.0172490 -4.0172161 -4.0171947 -4.0171809 -4.0171719
#>                 0.1984     0.1851     0.1726       0.1610       0.1501
#> (Intercept) -0.3618182 -5.5728049 -5.5656571 -26.05734839 -26.06065952
#> am           0.0000000  0.0000000  0.0000000   0.00000000   0.00000000
#> carb         0.0000000  0.0000000  0.0000000   0.00000000   0.00000000
#> cyl          0.0000000  0.0000000  0.0000000   0.00000000   0.00000000
#> disp         0.0000000  0.0000000  0.0000000   0.01965104   0.01965553
#> drat         2.2365126  1.8286655  1.8279618   2.19645842   2.19645822
#> gear         0.0000000  0.9303482  0.9301171   1.45404025   1.45410073
#> hp           0.0000000  0.0000000  0.0000000   0.00000000   0.00000000
#> qsec         1.4584906  1.5858337  1.5857001   2.75265933   2.75288230
#> vs           0.0000000  0.0000000  0.0000000  -2.66189479  -2.66238010
#> wt          -4.0171661 -3.7012353 -3.7016586  -5.85971512  -5.86024578
#>                   0.1400       0.1306       0.1218       0.1136       0.1059
#> (Intercept) -26.06375697 -26.06665454 -26.06936512 -26.07190078 -26.07427280
#> am            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> carb          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> cyl           0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> disp          0.01965972   0.01966365   0.01966732   0.01967075   0.01967397
#> drat          2.19645804   2.19645786   2.19645770   2.19645755   2.19645740
#> gear          1.45415731   1.45421023   1.45425974   1.45430605   1.45434938
#> hp            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> qsec          2.75309088   2.75328600   2.75346852   2.75363927   2.75379900
#> vs           -2.66283408  -2.66325877  -2.66365606  -2.66402770  -2.66437536
#> wt           -5.86074220  -5.86120658  -5.86164099  -5.86204737  -5.86242752
#>                   0.0988       0.0921       0.0859       0.0801       0.0747
#> (Intercept) -26.07649174 -10.63571222 -10.63874399 -10.64181606 -10.64491287
#> am            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> carb          0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> cyl           0.00000000  -0.97649526  -0.97619823  -0.97590490  -0.97561598
#> disp          0.01967697   0.02721462   0.02720984   0.02720526   0.02720091
#> drat          2.19645727   1.64771459   1.64788716   1.64805602   1.64822094
#> gear          1.45438991   0.65932853   0.65955214   0.65977503   0.65999645
#> hp            0.00000000   0.00000000   0.00000000   0.00000000   0.00000000
#> qsec          2.75394842   2.49221484   2.49217376   2.49213888   2.49210969
#> vs           -2.66470058  -3.63402602  -3.63345583  -3.63290257  -3.63236653
#> wt           -5.86278314  -6.24471895  -6.24429558  -6.24389532  -6.24351723
#>                   0.0697      0.0650       0.0606       0.0565       0.0527
#> (Intercept) -10.64802037 -10.6511260 -10.65421832 -10.65728740 -10.66032424
#> am            0.00000000   0.0000000   0.00000000   0.00000000   0.00000000
#> carb          0.00000000   0.0000000   0.00000000   0.00000000   0.00000000
#> cyl          -0.97533208  -0.9750537  -0.97478136  -0.97451539  -0.97425611
#> disp          0.02719675   0.0271928   0.02718904   0.02718547   0.02718208
#> drat          1.64838171   1.6485382   1.64869019   1.64883767   1.64898053
#> gear          0.66021574   0.6604323   0.66064561   0.66085523   0.66106077
#> hp            0.00000000   0.0000000   0.00000000   0.00000000   0.00000000
#> qsec          2.49208570   2.4920665   2.49205155   2.49204058   2.49203317
#> vs           -3.63184790  -3.6313468  -3.63086323  -3.63039719  -3.62994857
#> wt           -6.24316041  -6.2428240  -6.24250704  -6.24220876  -6.24192829
#>                   0.0492       0.0458      0.0428       0.0399       0.0372
#> (Intercept) -10.66332097 -11.84235883 -11.8399242 -11.83756164 -11.83526926
#> am            0.00000000   0.00000000   0.0000000   0.00000000   0.00000000
#> carb          0.00000000   0.33048018   0.3306758   0.33086553   0.33104944
#> cyl          -0.97400381  -1.09019850  -1.0905101  -1.09081229  -1.09110537
#> disp          0.02717886   0.03054753   0.0305524   0.03055712   0.03056169
#> drat          1.64911873   1.42231109   1.4220457   1.42178844   1.42153894
#> gear          0.66126188   0.59381827   0.5935712   0.59333162   0.59309920
#> hp            0.00000000   0.00000000   0.0000000   0.00000000   0.00000000
#> qsec          2.49202898   2.65595214   2.6560403   2.65612579   2.65620858
#> vs           -3.62951722  -3.63505717  -3.6354551  -3.63584090  -3.63621504
#> wt           -6.24166482  -6.73040739  -6.7309298  -6.73143630  -6.73192726
#>                   0.0347       0.0323      0.0302       0.0281       0.0262
#> (Intercept) -11.83304504 -11.83088706 -11.8287934 -11.82676230 -14.97955173
#> am            0.00000000   0.00000000   0.0000000   0.00000000  -0.37214604
#> carb          0.33122775   0.33140062   0.3315682   0.33173070   0.54753422
#> cyl          -1.09138960  -1.09166524  -1.0919325  -1.09219175  -1.56709458
#> disp          0.03056613   0.03057043   0.0305746   0.03057864   0.03267420
#> drat          1.42129701   1.42106243   1.4208350   1.42061445   1.24972366
#> gear          0.59287379   0.59265516   0.5924431   0.59223750   0.38799209
#> hp            0.00000000   0.00000000   0.0000000   0.00000000   0.01450945
#> qsec          2.65628880   2.65636653   2.6564418   2.65651483   3.07967344
#> vs           -3.63657783  -3.63692961  -3.6372707  -3.63760145  -4.56851341
#> wt           -6.73240319  -6.73286456  -6.7333118  -6.73374533  -7.54837149
#>                   0.0245       0.0228       0.0213       0.0198       0.0185
#> (Intercept) -14.98055415 -14.98154315 -14.98251890 -14.98348157 -14.98443135
#> am           -0.37217400  -0.37220160  -0.37222883  -0.37225570  -0.37228221
#> carb          0.54768705   0.54783783   0.54798659   0.54813334   0.54827812
#> cyl          -1.56744278  -1.56778630  -1.56812521  -1.56845956  -1.56878941
#> disp          0.03267641   0.03267858   0.03268073   0.03268285   0.03268494
#> drat          1.24952041   1.24931989   1.24912207   1.24892691   1.24873438
#> gear          0.38777665   0.38756412   0.38735444   0.38714759   0.38694351
#> hp            0.01451767   0.01452579   0.01453379   0.01454169   0.01454948
#> qsec          3.07993127   3.08018564   3.08043659   3.08068417   3.08092843
#> vs           -4.56916179  -4.56980146  -4.57043253  -4.57105513  -4.57166936
#> wt           -7.54890132  -7.54942403  -7.54993972  -7.55044847  -7.55095037
#>                   0.0173       0.0161       0.0150       0.0140       0.0131
#> (Intercept) -14.98536840 -14.98629289 -14.98720499 -14.98810487 -14.98899267
#> am           -0.37230837  -0.37233418  -0.37235964  -0.37238477  -0.37240957
#> carb          0.54842095   0.54856186   0.54870087   0.54883802   0.54897332
#> cyl          -1.56911483  -1.56943587  -1.56975260  -1.57006507  -1.57037334
#> disp          0.03268700   0.03268904   0.03269104   0.03269302   0.03269498
#> drat          1.24854444   1.24835705   1.24817219   1.24798981   1.24780988
#> gear          0.38674219   0.38654357   0.38634762   0.38615431   0.38596360
#> hp            0.01455717   0.01456475   0.01457223   0.01457961   0.01458690
#> qsec          3.08116940   3.08140713   3.08164167   3.08187306   3.08210133
#> vs           -4.57227533  -4.57287316  -4.57346296  -4.57404482  -4.57461887
#> wt           -7.55144553  -7.55193404  -7.55241597  -7.55289142  -7.55336048
#>                   0.0122       0.0114       0.0106       0.0099       0.0092
#> (Intercept) -14.98986858 -14.99073273 -14.99158530 -14.99242642 -14.99325627
#> am           -0.37243403  -0.37245817  -0.37248198  -0.37250548  -0.37252866
#> carb          0.54910680   0.54923849   0.54936841   0.54949658   0.54962302
#> cyl          -1.57067746  -1.57097749  -1.57127350  -1.57156552  -1.57185361
#> disp          0.03269690   0.03269880   0.03270068   0.03270253   0.03270435
#> drat          1.24763238   1.24745726   1.24728450   1.24711407   1.24694592
#> gear          0.38577545   0.38558984   0.38540672   0.38522607   0.38504784
#> hp            0.01459408   0.01460117   0.01460816   0.01461506   0.01462187
#> qsec          3.08232654   3.08254872   3.08276792   3.08298417   3.08319751
#> vs           -4.57518521  -4.57574393  -4.57629514  -4.57683894  -4.57737543
#> wt           -7.55382324  -7.55427977  -7.55473016  -7.55517449  -7.55561285
#>                   0.0086       0.0080       0.0075       0.0070       0.0065
#> (Intercept) -14.99407498 -14.99488271 -14.99567959 -14.99646578 -14.99724142
#> am           -0.37255154  -0.37257411  -0.37259637  -0.37261835  -0.37264002
#> carb          0.54974777   0.54987084   0.54999225   0.55011203   0.55023020
#> cyl          -1.57213783  -1.57241823  -1.57269486  -1.57296777  -1.57323701
#> disp          0.03270615   0.03270793   0.03270968   0.03271141   0.03271312
#> drat          1.24678004   1.24661639   1.24645494   1.24629567   1.24613854
#> gear          0.38487202   0.38469856   0.38452743   0.38435860   0.38419205
#> hp            0.01462858   0.01463520   0.01464174   0.01464819   0.01465455
#> qsec          3.08340799   3.08361563   3.08382049   3.08402259   3.08422197
#> vs           -4.57790472  -4.57842688  -4.57894203  -4.57945025  -4.57995164
#> wt           -7.55604531  -7.55647196  -7.55689288  -7.55730813  -7.55771779
#>                   0.0061       0.0057
#> (Intercept) -14.99800665 -14.99876161
#> am           -0.37266141  -0.37268251
#> carb          0.55034678   0.55046179
#> cyl          -1.57350263  -1.57376468
#> disp          0.03271480   0.03271646
#> drat          1.24598352   1.24583058
#> gear          0.38402773   0.38386563
#> hp            0.01466082   0.01466701
#> qsec          3.08441868   3.08461274
#> vs           -4.58044629  -4.58093428
#> wt           -7.55812195  -7.55852067
#> 
#> $iter
#>   [1]   0  10   4   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2
#>  [19]   3   3   3   3   3   3   3   3   3   3   3   3   3   3   3   2   1   1
#>  [37]  61  70  12   1   1   1   1   1   1   1   1   1   1  47   1 125   1   1
#>  [55]   1   1   1   1   1  95   1   1   1   1   1   1   1   1   1 126   1   1
#>  [73]   1   1   1   1   1 226   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.651843037 5.270927733 4.915684845 4.584384138 4.275411991 3.987263532
#>   [7] 3.718535315 3.467918480 3.234192381 3.016218639 2.812935598 2.623353153
#>  [13] 2.446547931 2.281658788 2.127882623 1.984470456 1.850723789 1.725991199
#>  [19] 1.609665169 1.501179124 1.400004676 1.305649046 1.217652670 1.135586956
#>  [25] 1.059052195 0.987675622 0.921109591 0.859029887 0.801134148 0.747140388
#>  [31] 0.696785626 0.649824607 0.606028603 0.565184304 0.527092773 0.491568484
#>  [37] 0.458438414 0.427541199 0.398726353 0.371853531 0.346791848 0.323419238
#>  [43] 0.301621864 0.281293560 0.262335315 0.244654793 0.228165879 0.212788262
#>  [49] 0.198447046 0.185072379 0.172599120 0.160966517 0.150117912 0.140000468
#>  [55] 0.130564905 0.121765267 0.113558696 0.105905220 0.098767562 0.092110959
#>  [61] 0.085902989 0.080113415 0.074714039 0.069678563 0.064982461 0.060602860
#>  [67] 0.056518430 0.052709277 0.049156848 0.045843841 0.042754120 0.039872635
#>  [73] 0.037185353 0.034679185 0.032341924 0.030162186 0.028129356 0.026233532
#>  [79] 0.024465479 0.022816588 0.021278826 0.019844705 0.018507238 0.017259912
#>  [85] 0.016096652 0.015011791 0.014000047 0.013056490 0.012176527 0.011355870
#>  [91] 0.010590522 0.009876756 0.009211096 0.008590299 0.008011341 0.007471404
#>  [97] 0.006967856 0.006498246 0.006060286 0.005651843
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 36
#> 
#> $loss
#>   [1] 864.78952 736.01430 628.27468 538.54347 464.20715 403.01065 353.00948
#>   [8] 312.52813 280.12395 254.55587 234.75723 219.81214 208.93512 201.45333
#>  [15] 196.79118 194.45705 193.97960 193.80575 181.04846 170.40959 161.57600
#>  [22] 154.28425 148.30716 143.44890 139.54080 136.43771 134.01484 132.16497
#>  [29] 130.79613 129.82948 129.19755 128.84266 128.71561 128.71352 128.71352
#>  [36] 128.71352 122.40667 116.33879 116.12625 116.12625 116.12625 116.12624
#>  [43] 116.12624 116.12624 116.12624 116.12624 116.12624 116.12624 116.12624
#>  [50] 113.57425 113.57424 101.87559 101.87557 101.87556 101.87554 101.87553
#>  [57] 101.87552 101.87551 101.87551  98.40339  98.40337  98.40335  98.40333
#>  [64]  98.40332  98.40330  98.40329  98.40327  98.40326  98.40325  98.01241
#>  [71]  98.01239  98.01236  98.01234  98.01232  98.01230  98.01229  98.01227
#>  [78]  97.33226  97.33223  97.33220  97.33217  97.33214  97.33212  97.33209
#>  [85]  97.33207  97.33204  97.33202  97.33200  97.33198  97.33196  97.33193
#>  [92]  97.33191  97.33189  97.33187  97.33186  97.33184  97.33182  97.33180
#>  [99]  97.33179  97.33177
#> 
#> $linear.predictors
#>     5.6518   5.2709   4.9157   4.5844   4.2754   3.9873   3.7185   3.4679
#> 1  21.5619 21.81392 22.04896 22.26815 22.47257 22.66322 22.84101 23.00682
#> 2  21.5619 21.67736 21.78503 21.88544 21.97909 22.06642 22.14787 22.22383
#> 3  21.5619 21.97459 22.35947 22.71840 23.05314 23.36533 23.65647 23.92799
#> 4  21.5619 21.49527 21.43312 21.37516 21.32111 21.27070 21.22368 21.17984
#> 5  21.5619 21.36405 21.17954 21.00746 20.84697 20.69731 20.55773 20.42755
#> 6  21.5619 21.50865 21.45899 21.41268 21.36949 21.32921 21.29164 21.25660
#> 7  21.5619 21.03736 20.54817 20.09195 19.66648 19.26968 18.89963 18.55451
#> 8  21.5619 21.21945 20.90008 20.60223 20.32446 20.06541 19.82381 19.59850
#> 9  21.5619 21.19267 20.84833 20.52719 20.22770 19.94839 19.68790 19.44497
#> 10 21.5619 20.31221 19.14675 18.05983 17.04617 16.10082 15.21919 14.39698
#> 11 21.5619 20.35452 19.22851 18.17840 17.19905 16.28571 15.43393 14.63955
#> 12 21.5619 22.03886 22.48367 22.89850 23.28537 23.64617 23.98265 24.29646
#> 13 21.5619 22.35216 23.08916 23.77648 24.41748 25.01529 25.57280 26.09273
#> 14 21.5619 22.23434 22.86145 23.44630 23.99173 24.50040 24.97479 25.41721
#> 15 21.5619 21.16054 20.78623 20.43714 20.11158 19.80797 19.52481 19.26074
#> 16 21.5619 21.15786 20.78105 20.42964 20.10191 19.79626 19.51122 19.24539
#> 17 21.5619 22.18078 22.75795 23.29622 23.79821 24.26637 24.70297 25.11015
#> 18 21.5619 22.07099 22.54577 22.98855 23.40149 23.78659 24.14574 24.48069
#> 19 21.5619 22.40679 23.19473 23.92957 24.61488 25.25400 25.85005 26.40593
#> 20 21.5619 21.51937 21.47969 21.44270 21.40819 21.37601 21.34600 21.31801
#> 21 21.5619 21.72823 21.88335 22.02802 22.16293 22.28876 22.40610 22.51553
#>      3.2342   3.0162   2.8129   2.6234   2.4465   2.2817   2.1279    1.9845
#> 1  23.16146 23.30567 23.44017 23.56560 23.68258 23.79167 23.89341 23.988293
#> 2  22.29467 22.36073 22.42235 22.47981 22.53340 22.58337 22.62998 22.673447
#> 3  24.18121 24.41736 24.63760 24.84300 25.03455 25.21319 25.37980 25.535170
#> 4  21.13895 21.10082 21.06525 21.03209 21.00116 20.97231 20.94541 20.920319
#> 5  20.30615 20.19294 20.08735 19.98888 19.89704 19.81140 19.73153 19.657035
#> 6  21.22393 21.19346 21.16504 21.13854 21.11382 21.09077 21.06927 21.049225
#> 7  18.23266 17.93250 17.65256 17.39150 17.14803 16.92096 16.70920 16.511718
#> 8  19.38838 19.19241 19.00966 18.83922 18.68026 18.53203 18.39378 18.264846
#> 9  19.21842 19.00713 18.81008 18.62632 18.45494 18.29510 18.14605 18.007033
#> 10 13.63018 12.91506 12.24814 11.62616 11.04611 10.50515 10.00064  9.530143
#> 11 13.89871 13.20780 12.56346 11.96254 11.40213 10.87948 10.39206  9.937488
#> 12 24.58911 24.86204 25.11658 25.35396 25.57534 25.78180 25.97435 26.153921
#> 13 26.57763 27.02984 27.45158 27.84489 28.21169 28.55378 28.87281 29.170333
#> 14 25.82981 26.21460 26.57346 26.90813 27.22025 27.51133 27.78279 28.035956
#> 15 19.01447 18.78479 18.57060 18.37084 18.18454 18.01080 17.84877 17.697657
#> 16 18.99747 18.76626 18.55064 18.34955 18.16201 17.98711 17.82399 17.671876
#> 17 25.48989 25.84404 26.17431 26.48233 26.76959 27.03748 27.28733 27.520330
#> 18 24.79306 25.08438 25.35606 25.60944 25.84574 26.06611 26.27163 26.463297
#> 19 26.92434 27.40782 27.85871 28.27921 28.67137 29.03710 29.37818 29.696271
#> 20 21.29191 21.26757 21.24487 21.22370 21.20395 21.18554 21.16837 21.152350
#> 21 22.61759 22.71277 22.80154 22.88432 22.96152 23.03352 23.10067 23.163291
#>       1.8507    1.7260    1.6097    1.5012    1.4000    1.3056    1.2177
#> 1  24.054800 24.052237 23.853798 23.668842 23.496352 23.335487 23.185463
#> 2  22.703914 22.702379 22.583513 22.472723 22.369401 22.273042 22.183177
#> 3  25.644078 25.644888 25.707615 25.766080 25.820605 25.871454 25.918877
#> 4  20.902733 20.905298 21.103888 21.288984 21.461605 21.622592 21.772730
#> 5  19.604823 19.608770 19.914374 20.199214 20.464856 20.712595 20.943637
#> 6  21.035173 21.038640 21.307097 21.557312 21.790664 22.008289 22.211247
#> 7  16.373292 16.372912 16.343469 16.316026 16.290433 16.266565 16.244306
#> 8  18.174473 18.174267 18.158300 18.143417 18.129538 18.116594 18.104522
#> 9  17.909594 17.910062 17.946352 17.980176 18.011720 18.041139 18.068574
#> 10  9.200352  9.201263  9.271808  9.337559  9.398878  9.456065  9.509397
#> 11  9.618862  9.619085  9.636367  9.652474  9.667496  9.681506  9.694571
#> 12 26.279789 26.281948 26.449142 26.604975 26.750306 26.885841 27.012243
#> 13 29.378880 29.379228 29.406148 29.431239 29.454638 29.476461 29.496813
#> 14 28.213410 28.216107 28.424961 28.619624 28.801167 28.970475 29.128372
#> 15 17.591738 17.588009 17.299236 17.030085 16.779073 16.544979 16.326662
#> 16 17.565250 17.564198 17.482747 17.406830 17.336030 17.270001 17.208423
#> 17 27.683651 27.684761 27.770772 27.850939 27.925702 27.995427 28.060452
#> 18 26.597644 26.595258 26.410494 26.238283 26.077680 25.927900 25.788215
#> 19 29.919234 29.916894 29.735656 29.566733 29.409195 29.262274 29.125255
#> 20 21.141125 21.135611 20.708670 20.310739 19.939627 19.593527 19.270753
#> 21 23.207185 23.208185 23.285595 23.357746 23.425033 23.487786 23.546309
#>       1.1356    1.0591    0.9877    0.9211    0.8590    0.8011    0.7471
#> 1  23.045551 22.915068 22.793380 22.679892 22.574054 22.475348 22.383295
#> 2  22.099369 22.021209 21.948317 21.880338 21.816940 21.757815 21.702675
#> 3  25.963104 26.004350 26.042816 26.078689 26.112145 26.143346 26.172444
#> 4  21.912748 22.043330 22.165111 22.278684 22.384603 22.483383 22.575506
#> 5  21.159108 21.360057 21.547462 21.722237 21.885232 22.037243 22.179008
#> 6  22.400526 22.577048 22.741674 22.895204 23.038386 23.171919 23.296452
#> 7  16.223546 16.204186 16.186131 16.169292 16.153589 16.138944 16.125285
#> 8  18.093264 18.082765 18.072973 18.063841 18.055325 18.047383 18.039976
#> 9  18.094161 18.118023 18.140277 18.161031 18.180387 18.198437 18.215272
#> 10  9.559135  9.605521  9.648781  9.689125  9.726750  9.761840  9.794564
#> 11  9.706756  9.718119  9.728717  9.738600  9.747817  9.756413  9.764430
#> 12 27.130125 27.240062 27.342590 27.438208 27.527381 27.610545 27.688103
#> 13 29.515793 29.533495 29.550003 29.565398 29.579756 29.593146 29.605634
#> 14 29.275627 29.412958 29.541033 29.660477 29.771870 29.875756 29.972640
#> 15 16.123059 15.933178 15.756095 15.590946 15.436928 15.293290 15.159333
#> 16 17.150994 17.097437 17.047488 17.000906 16.957464 16.916949 16.879166
#> 17 28.121095 28.177651 28.230396 28.279585 28.325459 28.368242 28.408141
#> 18 25.657945 25.536454 25.423151 25.317484 25.218939 25.127036 25.041327
#> 19 28.997471 28.878299 28.767159 28.663509 28.566845 28.476696 28.392623
#> 20 18.969733 18.689001 18.427189 18.183022 17.955311 17.742948 17.544897
#> 21 23.600888 23.651789 23.699259 23.743530 23.784817 23.823321 23.859231
#>       0.6968    0.6498    0.6060    0.5652    0.5271    0.4916   0.4584
#> 1  22.297447 22.217384 22.142717 22.131864 22.131847 22.131846 22.04695
#> 2  21.651251 21.603293 21.558567 21.552066 21.552056 21.552056 21.57640
#> 3  26.199581 26.224889 26.248491 26.251922 26.251927 26.251928 26.19890
#> 4  22.661420 22.741544 22.816267 22.827129 22.827145 22.827146 22.57374
#> 5  22.311219 22.434519 22.549509 22.566223 22.566249 22.566250 22.22429
#> 6  23.412592 23.520904 23.621917 23.636599 23.636622 23.636622 23.81541
#> 7  16.112548 16.100669 16.089590 16.087980 16.087977 16.087977 15.92478
#> 8  18.033068 18.026626 18.020618 18.019744 18.019743 18.019743 17.77601
#> 9  18.230971 18.245613 18.259268 18.261253 18.261256 18.261256 18.05891
#> 10  9.825083  9.853545  9.880089  9.883947  9.883953  9.883953 10.07917
#> 11  9.771907  9.778879  9.785382  9.786327  9.786328  9.786328 10.08317
#> 12 27.760435 27.827891 27.890801 27.899946 27.899959 27.899960 28.02386
#> 13 29.617280 29.628142 29.638271 29.639743 29.639745 29.639745 30.09427
#> 14 30.062994 30.147259 30.225845 30.237267 30.237285 30.237285 30.38148
#> 15 15.034404 14.917895 14.809238 14.793444 14.793420 14.793419 14.86298
#> 16 16.843928 16.811065 16.780418 16.775963 16.775956 16.775956 16.53285
#> 17 28.445351 28.480053 28.512416 28.517120 28.517127 28.517128 28.52832
#> 18 24.961394 24.886848 24.817327 24.807221 24.807206 24.807206 24.95377
#> 19 28.314216 28.241093 28.172898 28.162986 28.162971 28.162970 27.71597
#> 20 17.360193 17.187939 17.027293 17.003943 17.003908 17.003906 17.14686
#> 21 23.892720 23.923952 23.953079 23.957313 23.957319 23.957320 24.20192
#>      0.4275   0.3987   0.3719   0.3468   0.3234   0.3016   0.2813   0.2623
#> 1  21.88009 21.84271 21.84258 21.84250 21.84245 21.84242 21.84240 21.84238
#> 2  21.62406 21.63473 21.63473 21.63474 21.63474 21.63474 21.63474 21.63474
#> 3  26.09486 26.07156 26.07152 26.07149 26.07147 26.07146 26.07145 26.07145
#> 4  22.07649 21.96513 21.96491 21.96476 21.96467 21.96461 21.96457 21.96455
#> 5  21.55332 21.40306 21.40277 21.40258 21.40246 21.40238 21.40233 21.40229
#> 6  24.16664 24.24532 24.24555 24.24570 24.24580 24.24586 24.24590 24.24593
#> 7  15.60439 15.53263 15.53246 15.53234 15.53227 15.53223 15.53220 15.53218
#> 8  17.29754 17.19038 17.19013 17.18997 17.18987 17.18980 17.18975 17.18972
#> 9  17.66173 17.57278 17.57258 17.57245 17.57237 17.57231 17.57228 17.57225
#> 10 10.46247 10.54832 10.54853 10.54867 10.54876 10.54882 10.54885 10.54888
#> 11 10.66590 10.79642 10.79672 10.79692 10.79705 10.79713 10.79718 10.79722
#> 12 28.26724 28.32176 28.32192 28.32202 28.32208 28.32213 28.32216 28.32217
#> 13 30.98656 31.18640 31.18687 31.18717 31.18737 31.18750 31.18758 31.18763
#> 14 30.66476 30.72821 30.72839 30.72851 30.72859 30.72864 30.72868 30.72870
#> 15 14.99924 15.02975 15.02977 15.02978 15.02978 15.02979 15.02979 15.02979
#> 16 16.05554 15.94864 15.94838 15.94821 15.94810 15.94802 15.94798 15.94795
#> 17 28.55039 28.55533 28.55536 28.55538 28.55539 28.55540 28.55540 28.55540
#> 18 25.24128 25.30567 25.30578 25.30586 25.30591 25.30594 25.30596 25.30597
#> 19 26.83830 26.64172 26.64123 26.64092 26.64071 26.64058 26.64049 26.64044
#> 20 17.42705 17.48978 17.48984 17.48989 17.48991 17.48993 17.48994 17.48995
#> 21 24.68216 24.78972 24.78998 24.79015 24.79026 24.79033 24.79038 24.79041
#>      0.2447   0.2282   0.2128   0.1984   0.1851   0.1726   0.1610   0.1501
#> 1  21.84237 21.84237 21.84236 21.84236 21.68597 21.68614 21.42549 21.42541
#> 2  21.63474 21.63474 21.63474 21.63474 21.63022 21.63021 21.47275 21.47266
#> 3  26.07144 26.07144 26.07144 26.07144 26.11445 26.11450 25.30805 25.30789
#> 4  21.96453 21.96452 21.96451 21.96451 21.77967 21.77999 22.15065 22.15082
#> 5  21.40227 21.40226 21.40225 21.40224 21.52464 21.52499 21.51074 21.51081
#> 6  24.24594 24.24595 24.24596 24.24597 24.80610 24.80570 24.44535 24.44522
#> 7  15.53216 15.53216 15.53215 15.53215 15.36172 15.36197 14.51489 14.51472
#> 8  17.18971 17.18969 17.18969 17.18968 16.93731 16.93767 17.05772 17.05778
#> 9  17.57224 17.57223 17.57222 17.57222 17.38658 17.38687 17.86580 17.86592
#> 10 10.54889 10.54890 10.54891 10.54891 10.88829 10.88796 11.20292 11.20295
#> 11 10.79724 10.79726 10.79727 10.79727 10.96695 10.96654 10.67694 10.67683
#> 12 28.32218 28.32219 28.32220 28.32220 28.34301 28.34283 28.30791 28.30788
#> 13 31.18767 31.18769 31.18770 31.18771 30.55605 30.55565 30.92885 30.92891
#> 14 30.72871 30.72872 30.72872 30.72873 30.63188 30.63170 31.78850 31.78873
#> 15 15.02980 15.02980 15.02980 15.02980 14.26412 14.26426 13.29260 13.29244
#> 16 15.94793 15.94792 15.94791 15.94790 15.65774 15.65813 17.33252 17.33294
#> 17 28.55541 28.55541 28.55541 28.55541 28.41991 28.41992 28.29761 28.29760
#> 18 25.30598 25.30598 25.30599 25.30599 25.74270 25.74244 26.73680 26.73692
#> 19 26.64040 26.64038 26.64036 26.64035 27.17363 27.17407 26.35461 26.35451
#> 20 17.48996 17.48996 17.48996 17.48996 18.05758 18.05732 18.71769 18.71780
#> 21 24.79043 24.79044 24.79045 24.79045 24.87148 24.87115 23.41159 23.41125
#>      0.1400   0.1306   0.1218   0.1136   0.1059   0.0988   0.0921   0.0859
#> 1  21.42535 21.42528 21.42522 21.42517 21.42511 21.42507 21.58375 21.58373
#> 2  21.47259 21.47251 21.47245 21.47238 21.47232 21.47227 21.38699 21.38706
#> 3  25.30775 25.30762 25.30749 25.30738 25.30727 25.30716 25.63685 25.63683
#> 4  22.15098 22.15113 22.15127 22.15140 22.15152 22.15163 22.31749 22.31734
#> 5  21.51087 21.51093 21.51098 21.51103 21.51108 21.51113 21.30611 21.30614
#> 6  24.44510 24.44499 24.44488 24.44478 24.44468 24.44460 24.45769 24.45777
#> 7  14.51455 14.51440 14.51426 14.51413 14.51400 14.51389 14.04312 14.04335
#> 8  17.05782 17.05787 17.05791 17.05795 17.05799 17.05802 16.66477 16.66485
#> 9  17.86602 17.86612 17.86622 17.86630 17.86639 17.86646 17.34942 17.34950
#> 10 11.20297 11.20300 11.20302 11.20304 11.20306 11.20308 11.53209 11.53199
#> 11 10.67673 10.67663 10.67654 10.67646 10.67638 10.67631 10.86322 10.86324
#> 12 28.30785 28.30783 28.30781 28.30778 28.30776 28.30774 28.11110 28.11118
#> 13 30.92896 30.92901 30.92906 30.92910 30.92914 30.92918 30.71557 30.71560
#> 14 31.78894 31.78914 31.78933 31.78950 31.78966 31.78982 31.48593 31.48589
#> 15 13.29229 13.29214 13.29201 13.29189 13.29177 13.29166 13.62671 13.62669
#> 16 17.33334 17.33371 17.33406 17.33439 17.33469 17.33497 17.97244 17.97200
#> 17 28.29759 28.29757 28.29756 28.29755 28.29754 28.29753 28.35355 28.35354
#> 18 26.73703 26.73714 26.73723 26.73732 26.73741 26.73749 27.58453 27.58422
#> 19 26.35441 26.35432 26.35424 26.35416 26.35409 26.35402 26.59109 26.59108
#> 20 18.71791 18.71801 18.71811 18.71819 18.71828 18.71835 17.69601 17.69627
#> 21 23.41093 23.41064 23.41036 23.41010 23.40985 23.40962 23.52155 23.52171
#>      0.0801   0.0747   0.0697   0.0650   0.0606   0.0565   0.0527   0.0492
#> 1  21.58372 21.58369 21.58367 21.58364 21.58361 21.58358 21.58355 21.58352
#> 2  21.38712 21.38718 21.38723 21.38728 21.38732 21.38736 21.38740 21.38743
#> 3  25.63681 25.63679 25.63677 25.63674 25.63671 25.63668 25.63665 25.63662
#> 4  22.31720 22.31707 22.31695 22.31683 22.31673 22.31663 22.31653 22.31645
#> 5  21.30617 21.30620 21.30623 21.30626 21.30629 21.30633 21.30637 21.30640
#> 6  24.45785 24.45792 24.45799 24.45805 24.45811 24.45816 24.45821 24.45825
#> 7  14.04358 14.04379 14.04399 14.04419 14.04438 14.04456 14.04473 14.04489
#> 8  16.66493 16.66501 16.66509 16.66516 16.66524 16.66531 16.66539 16.66546
#> 9  17.34959 17.34968 17.34976 17.34985 17.34994 17.35002 17.35010 17.35019
#> 10 11.53189 11.53179 11.53169 11.53159 11.53150 11.53141 11.53132 11.53123
#> 11 10.86325 10.86325 10.86326 10.86326 10.86325 10.86325 10.86324 10.86323
#> 12 28.11126 28.11133 28.11140 28.11147 28.11154 28.11160 28.11166 28.11172
#> 13 30.71564 30.71567 30.71571 30.71574 30.71577 30.71581 30.71584 30.71588
#> 14 31.48587 31.48585 31.48583 31.48582 31.48581 31.48581 31.48581 31.48581
#> 15 13.62666 13.62663 13.62660 13.62657 13.62653 13.62650 13.62646 13.62642
#> 16 17.97158 17.97118 17.97080 17.97043 17.97009 17.96976 17.96945 17.96916
#> 17 28.35353 28.35352 28.35351 28.35350 28.35349 28.35348 28.35347 28.35345
#> 18 27.58390 27.58360 27.58330 27.58301 27.58272 27.58244 27.58217 27.58191
#> 19 26.59106 26.59105 26.59104 26.59103 26.59102 26.59101 26.59100 26.59099
#> 20 17.69653 17.69679 17.69706 17.69732 17.69759 17.69785 17.69811 17.69836
#> 21 23.52186 23.52200 23.52213 23.52225 23.52236 23.52245 23.52254 23.52263
#>      0.0458   0.0428   0.0399   0.0372   0.0347   0.0323   0.0302   0.0281
#> 1  21.83157 21.83175 21.83193 21.83211 21.83228 21.83244 21.83260 21.83276
#> 2  21.60265 21.60275 21.60285 21.60295 21.60304 21.60313 21.60322 21.60330
#> 3  25.45530 25.45523 25.45516 25.45510 25.45504 25.45498 25.45492 25.45486
#> 4  22.34876 22.34885 22.34895 22.34904 22.34912 22.34921 22.34929 22.34937
#> 5  21.30824 21.30820 21.30817 21.30813 21.30809 21.30806 21.30802 21.30799
#> 6  24.57672 24.57674 24.57677 24.57679 24.57682 24.57684 24.57686 24.57688
#> 7  13.82126 13.82099 13.82072 13.82046 13.82021 13.81996 13.81973 13.81950
#> 8  16.64079 16.64071 16.64063 16.64056 16.64048 16.64041 16.64034 16.64027
#> 9  17.36665 17.36658 17.36651 17.36644 17.36638 17.36631 17.36625 17.36619
#> 10 11.68156 11.68173 11.68189 11.68204 11.68219 11.68234 11.68248 11.68261
#> 11 10.86707 10.86708 10.86709 10.86710 10.86711 10.86712 10.86713 10.86714
#> 12 27.97916 27.97902 27.97889 27.97877 27.97864 27.97853 27.97841 27.97830
#> 13 30.84109 30.84113 30.84118 30.84122 30.84126 30.84130 30.84133 30.84137
#> 14 31.54478 31.54480 31.54482 31.54484 31.54485 31.54487 31.54489 31.54491
#> 15 13.61974 13.61979 13.61984 13.61989 13.61994 13.61998 13.62002 13.62006
#> 16 17.88377 17.88398 17.88419 17.88440 17.88459 17.88479 17.88497 17.88515
#> 17 28.25798 28.25794 28.25790 28.25786 28.25782 28.25778 28.25774 28.25770
#> 18 27.35393 27.35404 27.35415 27.35425 27.35435 27.35445 27.35454 27.35463
#> 19 26.76151 26.76162 26.76172 26.76183 26.76193 26.76202 26.76212 26.76221
#> 20 17.62732 17.62702 17.62673 17.62645 17.62617 17.62591 17.62565 17.62540
#> 21 23.43015 23.43003 23.42991 23.42980 23.42969 23.42959 23.42948 23.42938
#>      0.0262   0.0245   0.0228   0.0213   0.0198   0.0185   0.0173   0.0161
#> 1  21.60036 21.60032 21.60027 21.60022 21.60017 21.60013 21.60008 21.60003
#> 2  21.40015 21.40011 21.40007 21.40003 21.39999 21.39995 21.39992 21.39988
#> 3  25.40104 25.40105 25.40106 25.40107 25.40108 25.40109 25.40110 25.40110
#> 4  22.23685 22.23677 22.23669 22.23662 22.23655 22.23647 22.23640 22.23633
#> 5  21.23893 21.23888 21.23883 21.23877 21.23872 21.23867 21.23862 21.23857
#> 6  24.64913 24.64908 24.64903 24.64898 24.64893 24.64889 24.64884 24.64880
#> 7  13.61461 13.61443 13.61425 13.61408 13.61390 13.61373 13.61356 13.61340
#> 8  16.79699 16.79704 16.79709 16.79714 16.79719 16.79724 16.79729 16.79734
#> 9  17.65144 17.65157 17.65170 17.65182 17.65194 17.65207 17.65218 17.65230
#> 10 11.67405 11.67412 11.67419 11.67426 11.67433 11.67440 11.67447 11.67453
#> 11 10.89010 10.89014 10.89018 10.89022 10.89027 10.89031 10.89035 10.89039
#> 12 27.89369 27.89365 27.89362 27.89358 27.89354 27.89350 27.89347 27.89343
#> 13 30.69245 30.69233 30.69222 30.69210 30.69199 30.69188 30.69177 30.69167
#> 14 31.88524 31.88545 31.88566 31.88586 31.88607 31.88627 31.88647 31.88666
#> 15 13.96208 13.96222 13.96237 13.96251 13.96265 13.96278 13.96292 13.96305
#> 16 17.68566 17.68559 17.68552 17.68544 17.68537 17.68530 17.68523 17.68517
#> 17 28.14840 28.14836 28.14831 28.14827 28.14822 28.14818 28.14814 28.14810
#> 18 27.47933 27.47942 27.47951 27.47959 27.47968 27.47977 27.47985 27.47994
#> 19 26.93058 26.93066 26.93075 26.93083 26.93091 26.93099 26.93107 26.93115
#> 20 17.54154 17.54141 17.54127 17.54114 17.54101 17.54089 17.54076 17.54064
#> 21 23.42737 23.42740 23.42742 23.42744 23.42746 23.42749 23.42751 23.42753
#>      0.0150   0.0140   0.0131   0.0122   0.0114   0.0106   0.0099   0.0092
#> 1  21.59999 21.59995 21.59990 21.59986 21.59982 21.59978 21.59973 21.59969
#> 2  21.39984 21.39981 21.39977 21.39974 21.39970 21.39967 21.39964 21.39960
#> 3  25.40111 25.40112 25.40113 25.40114 25.40115 25.40115 25.40116 25.40117
#> 4  22.23626 22.23619 22.23613 22.23606 22.23599 22.23593 22.23586 22.23580
#> 5  21.23852 21.23847 21.23842 21.23838 21.23833 21.23828 21.23824 21.23819
#> 6  24.64875 24.64871 24.64866 24.64862 24.64858 24.64853 24.64849 24.64845
#> 7  13.61323 13.61307 13.61291 13.61275 13.61260 13.61244 13.61229 13.61214
#> 8  16.79738 16.79743 16.79747 16.79752 16.79756 16.79760 16.79765 16.79769
#> 9  17.65242 17.65253 17.65264 17.65276 17.65287 17.65297 17.65308 17.65319
#> 10 11.67460 11.67466 11.67472 11.67479 11.67485 11.67491 11.67497 11.67503
#> 11 10.89042 10.89046 10.89050 10.89054 10.89057 10.89061 10.89065 10.89068
#> 12 27.89340 27.89336 27.89333 27.89329 27.89326 27.89323 27.89319 27.89316
#> 13 30.69156 30.69146 30.69136 30.69126 30.69116 30.69106 30.69096 30.69087
#> 14 31.88685 31.88704 31.88723 31.88742 31.88760 31.88778 31.88796 31.88813
#> 15 13.96319 13.96332 13.96344 13.96357 13.96370 13.96382 13.96394 13.96406
#> 16 17.68510 17.68503 17.68497 17.68490 17.68484 17.68478 17.68471 17.68465
#> 17 28.14806 28.14802 28.14798 28.14794 28.14790 28.14786 28.14783 28.14779
#> 18 27.48002 27.48010 27.48018 27.48026 27.48034 27.48042 27.48049 27.48057
#> 19 26.93123 26.93131 26.93138 26.93146 26.93153 26.93160 26.93168 26.93175
#> 20 17.54051 17.54039 17.54027 17.54016 17.54004 17.53993 17.53981 17.53970
#> 21 23.42755 23.42757 23.42759 23.42761 23.42763 23.42765 23.42767 23.42769
#>      0.0086   0.0080   0.0075   0.0070   0.0065   0.0061   0.0057
#> 1  21.59965 21.59961 21.59957 21.59954 21.59950 21.59946 21.59942
#> 2  21.39957 21.39954 21.39951 21.39947 21.39944 21.39941 21.39938
#> 3  25.40118 25.40118 25.40119 25.40120 25.40121 25.40121 25.40122
#> 4  22.23574 22.23568 22.23561 22.23555 22.23550 22.23544 22.23538
#> 5  21.23815 21.23810 21.23806 21.23802 21.23798 21.23794 21.23790
#> 6  24.64841 24.64837 24.64833 24.64829 24.64825 24.64822 24.64818
#> 7  13.61199 13.61185 13.61170 13.61156 13.61142 13.61128 13.61115
#> 8  16.79773 16.79777 16.79781 16.79785 16.79789 16.79793 16.79797
#> 9  17.65329 17.65339 17.65349 17.65359 17.65369 17.65379 17.65389
#> 10 11.67509 11.67514 11.67520 11.67526 11.67531 11.67537 11.67542
#> 11 10.89072 10.89075 10.89078 10.89082 10.89085 10.89088 10.89091
#> 12 27.89313 27.89310 27.89307 27.89304 27.89301 27.89298 27.89295
#> 13 30.69077 30.69068 30.69059 30.69050 30.69041 30.69032 30.69023
#> 14 31.88831 31.88848 31.88865 31.88881 31.88898 31.88914 31.88930
#> 15 13.96418 13.96430 13.96441 13.96452 13.96464 13.96475 13.96486
#> 16 17.68459 17.68453 17.68447 17.68442 17.68436 17.68430 17.68425
#> 17 28.14775 28.14772 28.14768 28.14764 28.14761 28.14758 28.14754
#> 18 27.48064 27.48071 27.48079 27.48086 27.48093 27.48100 27.48107
#> 19 26.93182 26.93189 26.93195 26.93202 26.93209 26.93215 26.93222
#> 20 17.53959 17.53948 17.53938 17.53927 17.53917 17.53906 17.53896
#> 21 23.42771 23.42773 23.42775 23.42776 23.42778 23.42780 23.42782
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 21.0 22.8 21.4 18.1 24.4 16.4 17.3 15.2 10.4 14.7 32.4 30.4 33.9 13.3
#> [16] 19.2 27.3 26.0 30.4 15.8 21.4
#> 
#> $X
#>               am       carb         cyl        disp        drat       gear
#>  [1,]  0.9534626  1.3339459  0.05299989 -0.44591036  0.39041645  0.4082483
#>  [2,]  0.9534626  1.3339459  0.05299989 -0.44591036  0.39041645  0.4082483
#>  [3,]  0.9534626 -1.2126781 -1.05999788 -0.86316324  0.30187708  0.4082483
#>  [4,] -1.0488088 -1.2126781  0.05299989  0.34045084 -1.06162918 -1.0206207
#>  [5,] -1.0488088 -1.2126781  0.05299989  0.07565574 -1.62828113 -1.0206207
#>  [6,] -1.0488088 -0.3638034 -1.05999788 -0.55263081  0.01855111  0.4082483
#>  [7,] -1.0488088  0.4850713  1.16599767  0.48327971 -1.07933705 -1.0206207
#>  [8,] -1.0488088  0.4850713  1.16599767  0.48327971 -1.07933705 -1.0206207
#>  [9,] -1.0488088  0.4850713  1.16599767  0.48327971 -1.07933705 -1.0206207
#> [10,] -1.0488088  1.3339459  1.16599767  1.96131780 -1.20329216 -1.0206207
#> [11,] -1.0488088  1.3339459  1.16599767  1.80083592 -0.79601107 -1.0206207
#> [12,]  0.9534626 -1.2126781 -1.05999788 -1.09826919  0.70915817  0.4082483
#> [13,]  0.9534626 -0.3638034 -1.05999788 -1.12234147  2.21432741  0.4082483
#> [14,]  0.9534626 -1.2126781 -1.05999788 -1.15925230  0.95706840  0.4082483
#> [15,] -1.0488088  1.3339459  1.16599767  1.07866748  0.08938260 -1.0206207
#> [16,] -1.0488088 -0.3638034  1.16599767  1.47987217 -1.06162918 -1.0206207
#> [17,]  0.9534626 -1.2126781 -1.05999788 -1.09586196  0.70915817  0.4082483
#> [18,]  0.9534626 -0.3638034 -1.05999788 -0.76446689  1.32893374  1.8371173
#> [19,]  0.9534626 -0.3638034 -1.05999788 -0.96667405  0.16021409  1.8371173
#> [20,]  0.9534626  1.3339459  1.16599767  1.08669157  0.95706840  1.8371173
#> [21,]  0.9534626 -0.3638034 -1.05999788 -0.75885002  0.76228179  0.4082483
#>               hp        qsec         vs          wt
#>  [1,] -0.3800081 -0.98104084 -0.9534626 -0.44107655
#>  [2,] -0.3800081 -0.60002894 -0.9534626 -0.20205966
#>  [3,] -0.6454932  0.48177270  1.0488088 -0.72227289
#>  [4,] -0.3800081  1.04648676  1.0488088  0.11662953
#>  [5,] -0.4580920  1.57718191  1.0488088  0.34627321
#>  [6,] -1.1296131  1.42749866  1.0488088  0.09319650
#>  [7,]  0.7131659 -0.34148515 -0.9534626  0.91803910
#>  [8,]  0.7131659 -0.20540948 -0.9534626  0.59934991
#>  [9,]  0.7131659  0.06674188 -0.9534626  0.64621597
#> [10,]  1.2597529 -0.05572623 -0.9534626  2.18717192
#> [11,]  1.4940045 -0.32787759 -0.9534626  2.11312355
#> [12,] -1.0671461  1.06689811  1.0488088 -0.83475142
#> [13,] -1.2857809  0.42053864  1.0488088 -1.38308429
#> [14,] -1.0827628  1.35946082  1.0488088 -1.17687364
#> [15,]  1.7282561 -1.69543815 -0.9534626  0.70245524
#> [16,]  0.6350820 -0.57961759 -0.9534626  0.70714184
#> [17,] -1.0671461  0.67908243  1.0488088 -1.08314152
#> [18,] -0.6767268 -0.81775003 -0.9534626 -0.89099069
#> [19,] -0.3331578 -0.68167435  1.0488088 -1.47869104
#> [20,]  2.0249747 -2.31458249 -0.9534626  0.07445008
#> [21,] -0.3956249  0.47496891  1.0488088 -0.29110516
#> attr(,"center")
#>  [1]   0.5238095   2.4285714   5.9047619 215.5714286   3.6795238   3.7142857
#>  [7] 134.3333333  17.9019048   0.4761905   3.0905714
#> attr(,"scale")
#>  [1]   0.4994328   1.1780302   1.7969488 124.6246637   0.5647205   0.6998542
#>  [7]  64.0337213   1.4697704   0.4994328   1.0668702
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 30
#> 
#> $lambda.min
#> [1] 0.7471404
#> 
#> $null.dev
#> [1] 41.18045
#> 
#> $Bias
#> [1] 6.932373
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 7.897859