Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 38.385148 35.793919 31.561511 27.605454 24.372774 21.755224 19.659605
#>   [8] 18.005735 16.724700 15.757329 15.052873 14.567858 14.265097 14.112826
#>  [15] 14.083958 14.095573 13.677531 13.332433 13.292372 13.292372 13.173266
#>  [22] 13.040265 12.745495 12.360509 11.725776 10.920576 10.395567 10.139375
#>  [29] 10.035034  9.995733 11.333236 11.474574 11.617864 11.746276 11.854960
#>  [36] 19.430400 19.382783 19.648946 20.263883 21.428011 20.847882 21.434052
#>  [43] 21.605587 22.097405 22.238989 22.279456 24.794926 23.872105 23.338460
#>  [50] 23.215777 26.119044 31.833223 33.003308 33.005283 32.969630 32.660950
#>  [57] 32.577684 32.524714 31.866827 37.414394 37.414356 37.414414 37.414542
#>  [64] 37.151778 37.547079 37.546827 37.546529 37.554855 37.572757 38.614261
#>  [71] 38.528657 38.934245 38.952004 38.932256 40.024879 40.025457 40.016147
#>  [78] 40.015739 40.015408 40.015151 40.014960 40.015041 40.024898 40.025255
#>  [85] 40.025893 40.092673 40.092428 40.126933 40.126784 40.126661 40.126560
#>  [92] 40.126482 40.127553 40.360987 40.397772 40.388249 40.389389 40.390523
#>  [99] 40.391650 40.392769
#> 
#> $cvse
#>   [1] 12.641372 12.709938 12.019296 11.042194 10.155368  9.349309  8.616649
#>   [8]  7.951748  7.350350  6.809287  6.326216  5.899367  5.527294  5.208623
#>  [15]  4.941800  4.716010  4.478198  4.275465  4.244596  4.244596  4.240061
#>  [22]  4.232286  4.173872  4.114801  3.951078  3.719129  3.534197  3.371061
#>  [29]  3.245205  3.122985  3.193306  3.211935  3.234458  3.254326  3.269763
#>  [36]  8.574643  8.570766  8.578703  8.539347  8.650578  8.029766  7.951519
#>  [43]  7.919720  7.890061  7.887672  7.880166 10.041628 10.065361 10.047353
#>  [50] 10.048882 10.309076 16.614673 16.751149 16.753122 16.758654 16.729458
#>  [57] 16.739417 16.745366 16.728051 17.246646 17.247684 17.248677 17.249629
#>  [64] 17.258394 17.221666 17.222486 17.223278 17.224043 17.225670 18.243312
#>  [71] 18.249504 18.219996 18.219014 18.221893 18.164608 18.165409 18.165956
#>  [78] 18.166703 18.167423 18.168116 18.168784 18.169429 18.170874 18.171468
#>  [85] 18.172072 18.165798 18.166251 18.163841 18.164258 18.164659 18.165044
#>  [92] 18.165415 18.166895 18.415730 18.419592 18.421046 18.422076 18.423086
#>  [99] 18.424076 18.425047
#> 
#> $fold
#>  [1]  5  7  7  1 10  3  8  5  9  3  6  1 10  9  2  2  8  6  4  1  4
#> 
#> $lambda
#>   [1] 4.989781388 4.653486822 4.339857385 4.047365522 3.774586632 3.520192126
#>   [7] 3.282942959 3.061683591 2.855336364 2.662896250 2.483425955 2.316051357
#>  [13] 2.159957247 2.014383358 1.878620661 1.752007916 1.633928447 1.523807138
#>  [19] 1.421107637 1.325329738 1.236006950 1.152704218 1.075015812 1.002563344
#>  [25] 0.934993930 0.871978468 0.813210038 0.758402403 0.707288620 0.659619734
#>  [31] 0.615163571 0.573703606 0.535037903 0.498978139 0.465348682 0.433985739
#>  [37] 0.404736552 0.377458663 0.352019213 0.328294296 0.306168359 0.285533636
#>  [43] 0.266289625 0.248342595 0.231605136 0.215995725 0.201438336 0.187862066
#>  [49] 0.175200792 0.163392845 0.152380714 0.142110764 0.132532974 0.123600695
#>  [55] 0.115270422 0.107501581 0.100256334 0.093499393 0.087197847 0.081321004
#>  [61] 0.075840240 0.070728862 0.065961973 0.061516357 0.057370361 0.053503790
#>  [67] 0.049897814 0.046534868 0.043398574 0.040473655 0.037745866 0.035201921
#>  [73] 0.032829430 0.030616836 0.028553364 0.026628962 0.024834260 0.023160514
#>  [79] 0.021599572 0.020143834 0.018786207 0.017520079 0.016339284 0.015238071
#>  [85] 0.014211076 0.013253297 0.012360069 0.011527042 0.010750158 0.010025633
#>  [91] 0.009349939 0.008719785 0.008132100 0.007584024 0.007072886 0.006596197
#>  [97] 0.006151636 0.005737036 0.005350379 0.004989781
#> 
#> $fit
#> $beta
#>              4.9898     4.6535    4.3399    4.0474    3.7746   3.5202    3.2829
#> (Intercept) 19.8619 21.5749810 23.253215 24.818341 26.277983 27.63925 28.908772
#> am           0.0000  0.0000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> carb         0.0000  0.0000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> cyl          0.0000 -0.2767277 -0.547827 -0.800655 -1.036443 -1.25634 -1.461417
#> disp         0.0000  0.0000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> drat         0.0000  0.0000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> gear         0.0000  0.0000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> hp           0.0000  0.0000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> qsec         0.0000  0.0000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> vs           0.0000  0.0000000  0.000000  0.000000  0.000000  0.00000  0.000000
#> wt           0.0000  0.0000000  0.000000  0.000000  0.000000  0.00000  0.000000
#>                3.0617    2.8553    2.6629    2.4834    2.3161    2.1600
#> (Intercept) 30.092732 31.196898 32.226646 33.186993 34.082616 34.917877
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -1.652672 -1.831037 -1.997381 -2.152514 -2.297192 -2.432119
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                2.0144    1.8786    1.7520    1.6339    1.5238    1.4211
#> (Intercept) 35.696844 36.423311 37.100817 37.604518 37.604518 37.604518
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> cyl         -2.557952 -2.675304 -2.784747 -2.866114 -2.866114 -2.866114
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#> wt           0.000000  0.000000  0.000000  0.000000  0.000000  0.000000
#>                1.3253    1.2360     1.1527     1.0750    1.0026    0.9350
#> (Intercept) 37.604518 37.604518 37.7187211 38.3139102 38.869078 39.386664
#> am           0.000000  0.000000  0.0000000  0.0000000  0.000000  0.000000
#> carb         0.000000  0.000000  0.0000000  0.0000000  0.000000  0.000000
#> cyl         -2.866114 -2.866114 -2.8204738 -2.5789755 -2.353623 -2.143691
#> disp         0.000000  0.000000  0.0000000  0.0000000  0.000000  0.000000
#> drat         0.000000  0.000000  0.0000000  0.0000000  0.000000  0.000000
#> gear         0.000000  0.000000  0.0000000  0.0000000  0.000000  0.000000
#> hp           0.000000  0.000000  0.0000000  0.0000000  0.000000  0.000000
#> qsec         0.000000  0.000000  0.0000000  0.0000000  0.000000  0.000000
#> vs           0.000000  0.000000  0.0000000  0.0000000  0.000000  0.000000
#> wt           0.000000  0.000000 -0.1209855 -0.7583832 -1.353097 -1.907238
#>                0.8720    0.8132    0.7584    0.7073    0.6596      0.6152
#> (Intercept) 39.869532 40.185476 40.185295 40.185199 40.185148 40.14757842
#> am           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> carb         0.000000  0.000000  0.000000  0.000000  0.000000 -0.04680316
#> cyl         -1.947676 -1.819091 -1.818977 -1.818917 -1.818884 -1.78953874
#> disp         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> drat         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> gear         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> hp           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> qsec         0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> vs           0.000000  0.000000  0.000000  0.000000  0.000000  0.00000000
#> wt          -2.424522 -2.763609 -2.763769 -2.763854 -2.763899 -2.76638183
#>                 0.5737     0.5350     0.4990     0.4653    0.4340     0.4047
#> (Intercept) 40.0926504 40.0416267 39.9938727 39.9493353 39.907800 39.8690632
#> am           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> carb        -0.1152137 -0.1788089 -0.2382867 -0.2937577 -0.345490 -0.3937359
#> cyl         -1.7466361 -1.7067772 -1.6694777 -1.6346907 -1.602248 -1.5719922
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> drat         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> gear         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> hp           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> qsec         0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> vs           0.0000000  0.0000000  0.0000000  0.0000000  0.000000  0.0000000
#> wt          -2.7700240 -2.7733762 -2.7765412 -2.7794934 -2.782247 -2.7848142
#>                0.3775     0.3520     0.3283     0.3062     0.2855     0.2663
#> (Intercept) 39.832937 39.7992466 39.7678263 39.7597141 39.7596017 39.7595510
#> am           0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> carb        -0.438730 -0.4806918 -0.5198254 -0.5299637 -0.5300615 -0.5300999
#> cyl         -1.543775 -1.5174603 -1.4929188 -1.4865783 -1.4864956 -1.4864591
#> disp         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> drat         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> gear         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> hp           0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> qsec         0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> vs           0.000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> wt          -2.787209 -2.7894421 -2.7915248 -2.7920399 -2.7920749 -2.7920945
#>                0.2483     0.2316    0.2160     0.2014     0.1879     0.1752
#> (Intercept) 39.759527 39.5505615  7.690920  7.2695774 -1.0642739 -1.0672187
#> am           0.000000  0.1233182  3.437155  3.3519867  3.0101322  3.0104052
#> carb        -0.530116 -0.5469780 -0.400825 -0.4132526 -0.5186533 -0.5185843
#> cyl         -1.486442 -1.4802028  0.000000  0.0000000  0.0000000  0.0000000
#> disp         0.000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> drat         0.000000  0.0000000  0.000000  0.1758854  1.3812509  1.3811581
#> gear         0.000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> hp           0.000000  0.0000000  0.000000  0.0000000  0.0000000  0.0000000
#> qsec         0.000000  0.0000000  1.217198  1.2040366  1.5486268  1.5488149
#> vs           0.000000  0.0000000  0.000000  0.0000000 -2.3757328 -2.3762688
#> wt          -2.792105 -2.7413399 -3.024687 -2.9941550 -3.1653175 -3.1653637
#>                 0.1634     0.1524     0.1421     0.1325      0.1236      0.1153
#> (Intercept) -1.0695759 -1.0714621 -1.0729708 -1.0741771  4.14828345  4.14409599
#> am           3.0106246  3.0108007  3.0109420  3.0110553  3.57614692  3.57639646
#> carb        -0.5185289 -0.5184846 -0.5184491 -0.5184207  0.00000000  0.00000000
#> cyl          0.0000000  0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> disp         0.0000000  0.0000000  0.0000000  0.0000000  0.00000000  0.00000000
#> drat         1.3810834  1.3810234  1.3809752  1.3809364  1.79562999  1.79607659
#> gear         0.0000000  0.0000000  0.0000000  0.0000000 -1.22386019 -1.22418133
#> hp           0.0000000  0.0000000  0.0000000  0.0000000 -0.01169364 -0.01168837
#> qsec         1.5489654  1.5490858  1.5491822  1.5492592  1.40359814  1.40379464
#> vs          -2.3766971 -2.3770392 -2.3773124 -2.3775306 -1.95213617 -1.95247938
#> wt          -3.1654003 -3.1654294 -3.1654524 -3.1654708 -3.07786983 -3.07801787
#>                  0.1075      0.1003      0.0935      0.0872      0.0813
#> (Intercept)  4.14072743  4.13800948  4.13580261  4.13399498  4.13249922
#> am           3.57658439  3.57673108  3.57685004  3.57695001  3.57703648
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> drat         1.79638191  1.79658809  1.79672619  1.79681846  1.79688047
#> gear        -1.22441027 -1.22457776 -1.22470435 -1.22480354 -1.22488401
#> hp          -0.01168408 -0.01168054 -0.01167757 -0.01167504 -0.01167288
#> qsec         1.40396081  1.40410172  1.40422166  1.40432417  1.40441215
#> vs          -1.95277734 -1.95303377 -1.95325357 -1.95344179 -1.95360308
#> wt          -3.07815282 -3.07827574 -3.07838706 -3.07848713 -3.07857635
#>                  0.0758      0.0707      0.0660     0.0615      0.0574
#> (Intercept)  4.13124832  4.13019152  4.12929061  4.1285168  4.12784823
#> am           3.57711281  3.57718105  3.57724244  3.5772977  3.57734752
#> carb         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> drat         1.79692283  1.79695262  1.79697443  1.7969911  1.79700458
#> gear        -1.22495128 -1.22500881 -1.22505878 -1.2251026 -1.22514125
#> hp          -0.01167103 -0.01166942 -0.01166802 -0.0116668 -0.01166574
#> qsec         1.40448795  1.40455346  1.40461025  1.4046596  1.40470245
#> vs          -1.95374157 -1.95386075 -1.95396353 -1.9540523 -1.95412922
#> wt          -3.07865531 -3.07872473 -3.07878545 -3.0788383 -3.07888422
#>                  0.0535      0.0499      0.0465     0.0434      0.0405
#> (Intercept)  4.12726812  4.12676331  4.12632323  4.1259392  4.12560413
#> am           3.57739218  3.57743209  3.57746762  3.5774991  3.57752697
#> carb         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> disp         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> drat         1.79701581  1.79702548  1.79703396  1.7970415  1.79704816
#> gear        -1.22517536 -1.22520547 -1.22523201 -1.2252554 -1.22527588
#> hp          -0.01166482 -0.01166401 -0.01166331 -0.0116627 -0.01166217
#> qsec         1.40473979  1.40477232  1.40480065  1.4048253  1.40484684
#> vs          -1.95419583 -1.95425360 -1.95430372 -1.9543472 -1.95438497
#> wt          -3.07892399 -3.07895840 -3.07898815 -3.0790139 -3.07903613
#>                   0.0377       0.0352       0.0328       0.0306       0.0286
#> (Intercept)  3.004011853  3.001215798  2.998728699  2.996523928  2.994576378
#> am           3.630778638  3.631350201  3.631855006  3.632299143  3.632688313
#> carb         0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> cyl          0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> disp         0.004574107  0.004575611  0.004576454  0.004576746  0.004576581
#> drat         1.805892899  1.805646148  1.805442357  1.805276208  1.805142895
#> gear        -1.011540828 -1.011781456 -1.012024115 -1.012265631 -1.012503414
#> hp          -0.014797893 -0.014791425 -0.014785259 -0.014779411 -0.014773893
#> qsec         1.423905954  1.424150671  1.424370601  1.424567641  1.424743611
#> vs          -1.840061292 -1.840264050 -1.840467992 -1.840670579 -1.840869734
#> wt          -3.296359813 -3.296755625 -3.297084851 -3.297355591 -3.297575175
#>                   0.0266       0.0248       0.0232       0.0216       0.0201
#> (Intercept)  2.992862479  2.991360192  2.990048983  2.757845395  2.756947156
#> am           3.633027839  3.633322668  3.633577379  3.779350032  3.779622285
#> carb         0.000000000  0.000000000  0.000000000  0.172340441  0.172534240
#> cyl          0.000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> disp         0.004576044  0.004575206  0.004574131  0.007264724  0.007269354
#> drat         1.805038092  1.804957915  1.804898890  1.839984298  1.839976005
#> gear        -1.012735379 -1.012959882 -1.013175661 -1.193542607 -1.193727820
#> hp          -0.014768709 -0.014763860 -0.014759341 -0.017752326 -0.017755486
#> qsec         1.424900250  1.425039207  1.425162041  1.464266106  1.464350346
#> vs          -1.841063788 -1.841251425 -1.841431633 -1.657178619 -1.656946500
#> wt          -3.297750213 -3.297886649 -3.297989810 -3.525258211 -3.525649025
#>                   0.0188       0.0175       0.0163       0.0152       0.0142
#> (Intercept)  2.756081801  2.755248140  2.754445027  2.753671358  0.355994866
#> am           3.779884543  3.780137169  3.780380516  3.780614920  3.940625659
#> carb         0.172720904  0.172900695  0.173073865  0.173240657  0.182040834
#> cyl          0.000000000  0.000000000  0.000000000  0.000000000  0.209254976
#> disp         0.007273813  0.007278108  0.007282244  0.007286228  0.006424823
#> drat         1.839968019  1.839960330  1.839952926  1.839945796  1.942907346
#> gear        -1.193906244 -1.194078126 -1.194243703 -1.194403204 -1.191809906
#> hp          -0.017758528 -0.017761457 -0.017764279 -0.017766995 -0.018979878
#> qsec         1.464431499  1.464509678  1.464584990  1.464657540  1.510615679
#> vs          -1.656722950 -1.656507654 -1.656300306 -1.656100612 -1.485189660
#> wt          -3.526025462 -3.526388050 -3.526737295 -3.527073684 -3.492347075
#>                  0.0133       0.0124       0.0115       0.0108       0.0100
#> (Intercept)  0.35553376  0.355153470  0.354848306  0.354612881  0.354442108
#> am           3.94086402  3.941090524  3.941305678  3.941509972  3.941703878
#> carb         0.18221880  0.182391480  0.182558971  0.182721359  0.182878739
#> cyl          0.20922899  0.209197703  0.209161589  0.209121103  0.209076670
#> disp         0.00642955  0.006434152  0.006438629  0.006442983  0.006447215
#> drat         1.94287278  1.942836832  1.942799692  1.942761523  1.942722480
#> gear        -1.19196019 -1.192107270 -1.192251113 -1.192391702 -1.192529026
#> hp          -0.01898294 -0.018985875 -0.018988679 -0.018991358 -0.018993918
#> qsec         1.51068036  1.510742158  1.510801172  1.510857515  1.510911291
#> vs          -1.48494872 -1.484721787 -1.484508181 -1.484307255 -1.484118388
#> wt          -3.49272050 -3.493083919 -3.493437424 -3.493781125 -3.494115142
#>                   0.0093       0.0087       0.0081       0.0076       0.0071
#> (Intercept)  0.354331179  0.354275555  0.354270950  0.354313324  0.354398867
#> am           3.941887849  3.942062321  3.942227713  3.942384431  3.942532861
#> carb         0.183031205  0.183178855  0.183321788  0.183460103  0.183593902
#> cyl          0.209028687  0.208977530  0.208923547  0.208867067  0.208808396
#> disp         0.006451326  0.006455319  0.006459195  0.006462956  0.006466604
#> drat         1.942682705  1.942642334  1.942601491  1.942560291  1.942518842
#> gear        -1.192663082 -1.192793873 -1.192921407 -1.193045698 -1.193166767
#> hp          -0.018996364 -0.018998700 -0.019000930 -0.019003057 -0.019005087
#> qsec         1.510962600  1.511011540  1.511058206  1.511102689  1.511145076
#> vs          -1.483940986 -1.483774479 -1.483618323 -1.483471995 -1.483334997
#> wt          -3.494439603 -3.494754644 -3.495060408 -3.495357044 -3.495644709
#>                   0.0066       0.0062       0.0057       0.0054       0.0050
#> (Intercept)  0.354523991  0.354685314  0.354879656  0.355104025  0.355355612
#> am           3.942673377  3.942806338  3.942932088  3.943050958  3.943163265
#> carb         0.183723286  0.183848357  0.183969218  0.184085969  0.184198714
#> cyl          0.208747819  0.208685605  0.208622003  0.208557244  0.208491544
#> disp         0.006470141  0.006473568  0.006476889  0.006480104  0.006483217
#> drat         1.942477242  1.942435584  1.942393952  1.942352423  1.942311068
#> gear        -1.193284635 -1.193399333 -1.193510891 -1.193619345 -1.193724733
#> hp          -0.019007024 -0.019008869 -0.019010628 -0.019012304 -0.019013901
#> qsec         1.511185453  1.511223903  1.511260503  1.511295332  1.511328462
#> vs          -1.483206852 -1.483087104 -1.482975317 -1.482871073 -1.482773975
#> wt          -3.495923562 -3.496193768 -3.496455493 -3.496708910 -3.496954190
#> 
#> $iter
#>   [1]   0   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1  18  25  25  24  24  12   1   1   1   9   9   8   8   8   8
#>  [37]   8   8   8   8   4   1   1   1  40 115  26  76   1   1   1   1   1  37
#>  [55]   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1  46   1
#>  [73]   1   1   1   1   1   1  99   1   1   1   1   1  88   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 4.989781388 4.653486822 4.339857385 4.047365522 3.774586632 3.520192126
#>   [7] 3.282942959 3.061683591 2.855336364 2.662896250 2.483425955 2.316051357
#>  [13] 2.159957247 2.014383358 1.878620661 1.752007916 1.633928447 1.523807138
#>  [19] 1.421107637 1.325329738 1.236006950 1.152704218 1.075015812 1.002563344
#>  [25] 0.934993930 0.871978468 0.813210038 0.758402403 0.707288620 0.659619734
#>  [31] 0.615163571 0.573703606 0.535037903 0.498978139 0.465348682 0.433985739
#>  [37] 0.404736552 0.377458663 0.352019213 0.328294296 0.306168359 0.285533636
#>  [43] 0.266289625 0.248342595 0.231605136 0.215995725 0.201438336 0.187862066
#>  [49] 0.175200792 0.163392845 0.152380714 0.142110764 0.132532974 0.123600695
#>  [55] 0.115270422 0.107501581 0.100256334 0.093499393 0.087197847 0.081321004
#>  [61] 0.075840240 0.070728862 0.065961973 0.061516357 0.057370361 0.053503790
#>  [67] 0.049897814 0.046534868 0.043398574 0.040473655 0.037745866 0.035201921
#>  [73] 0.032829430 0.030616836 0.028553364 0.026628962 0.024834260 0.023160514
#>  [79] 0.021599572 0.020143834 0.018786207 0.017520079 0.016339284 0.015238071
#>  [85] 0.014211076 0.013253297 0.012360069 0.011527042 0.010750158 0.010025633
#>  [91] 0.009349939 0.008719785 0.008132100 0.007584024 0.007072886 0.006596197
#>  [97] 0.006151636 0.005737036 0.005350379 0.004989781
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 21
#> 
#> $loss
#>   [1] 723.5295 628.0597 543.9235 473.8346 415.7550 367.9266 328.8329 297.1675
#>   [9] 271.8054 251.7787 236.2555 224.5216 215.9642 210.0585 206.3555 204.4718
#>  [17] 204.0531 204.0531 204.0531 204.0531 204.0531 198.8190 175.1117 158.8548
#>  [25] 148.8009 143.8502 142.9283 142.9283 142.9283 142.9283 140.7231 137.8661
#>  [33] 135.6006 133.8221 132.4598 131.4472 130.7271 130.2506 129.9758 129.8671
#>  [41] 129.8622 129.8622 129.8622 129.8622 129.2908 113.0953 112.4295 105.8669
#>  [49] 105.8669 105.8669 105.8669 105.8669 105.8669 101.5602 101.5602 101.5602
#>  [57] 101.5602 101.5602 101.5602 101.5602 101.5602 101.5602 101.5602 101.5602
#>  [65] 101.5602 101.5602 101.5602 101.5602 101.5602 101.5602 101.1008 101.1008
#>  [73] 101.1007 101.1007 101.1007 101.1007 101.1007 101.1007 100.9474 100.9474
#>  [81] 100.9474 100.9473 100.9473 100.9473 100.8371 100.8371 100.8371 100.8371
#>  [89] 100.8370 100.8370 100.8370 100.8370 100.8370 100.8370 100.8370 100.8370
#>  [97] 100.8370 100.8370 100.8370 100.8370
#> 
#> $linear.predictors
#>     4.9898   4.6535   4.3399   4.0474   3.7746   3.5202   3.2829   3.0617
#> 1  19.8619 19.91461 19.96625 20.01441 20.05932 20.10121 20.14027 20.17670
#> 2  19.8619 20.46807 21.06191 21.61572 22.13221 22.61389 23.06310 23.48204
#> 3  19.8619 19.91461 19.96625 20.01441 20.05932 20.10121 20.14027 20.17670
#> 4  19.8619 19.36116 18.87060 18.41310 17.98644 17.58853 17.21744 16.87136
#> 5  19.8619 19.91461 19.96625 20.01441 20.05932 20.10121 20.14027 20.17670
#> 6  19.8619 19.36116 18.87060 18.41310 17.98644 17.58853 17.21744 16.87136
#> 7  19.8619 20.46807 21.06191 21.61572 22.13221 22.61389 23.06310 23.48204
#> 8  19.8619 20.46807 21.06191 21.61572 22.13221 22.61389 23.06310 23.48204
#> 9  19.8619 19.91461 19.96625 20.01441 20.05932 20.10121 20.14027 20.17670
#> 10 19.8619 19.36116 18.87060 18.41310 17.98644 17.58853 17.21744 16.87136
#> 11 19.8619 19.36116 18.87060 18.41310 17.98644 17.58853 17.21744 16.87136
#> 12 19.8619 19.36116 18.87060 18.41310 17.98644 17.58853 17.21744 16.87136
#> 13 19.8619 20.46807 21.06191 21.61572 22.13221 22.61389 23.06310 23.48204
#> 14 19.8619 20.46807 21.06191 21.61572 22.13221 22.61389 23.06310 23.48204
#> 15 19.8619 19.36116 18.87060 18.41310 17.98644 17.58853 17.21744 16.87136
#> 16 19.8619 19.36116 18.87060 18.41310 17.98644 17.58853 17.21744 16.87136
#> 17 19.8619 20.46807 21.06191 21.61572 22.13221 22.61389 23.06310 23.48204
#> 18 19.8619 19.36116 18.87060 18.41310 17.98644 17.58853 17.21744 16.87136
#> 19 19.8619 19.91461 19.96625 20.01441 20.05932 20.10121 20.14027 20.17670
#> 20 19.8619 19.36116 18.87060 18.41310 17.98644 17.58853 17.21744 16.87136
#> 21 19.8619 20.46807 21.06191 21.61572 22.13221 22.61389 23.06310 23.48204
#>      2.8553   2.6629   2.4834   2.3161   2.1600   2.0144   1.8786   1.7520
#> 1  20.21067 20.24236 20.27191 20.29947 20.32517 20.34913 20.37149 20.39233
#> 2  23.87275 24.23712 24.57694 24.89385 25.18940 25.46504 25.72209 25.96183
#> 3  20.21067 20.24236 20.27191 20.29947 20.32517 20.34913 20.37149 20.39233
#> 4  16.54860 16.24760 15.96688 15.70508 15.46093 15.23323 15.02088 14.82284
#> 5  20.21067 20.24236 20.27191 20.29947 20.32517 20.34913 20.37149 20.39233
#> 6  16.54860 16.24760 15.96688 15.70508 15.46093 15.23323 15.02088 14.82284
#> 7  23.87275 24.23712 24.57694 24.89385 25.18940 25.46504 25.72209 25.96183
#> 8  23.87275 24.23712 24.57694 24.89385 25.18940 25.46504 25.72209 25.96183
#> 9  20.21067 20.24236 20.27191 20.29947 20.32517 20.34913 20.37149 20.39233
#> 10 16.54860 16.24760 15.96688 15.70508 15.46093 15.23323 15.02088 14.82284
#> 11 16.54860 16.24760 15.96688 15.70508 15.46093 15.23323 15.02088 14.82284
#> 12 16.54860 16.24760 15.96688 15.70508 15.46093 15.23323 15.02088 14.82284
#> 13 23.87275 24.23712 24.57694 24.89385 25.18940 25.46504 25.72209 25.96183
#> 14 23.87275 24.23712 24.57694 24.89385 25.18940 25.46504 25.72209 25.96183
#> 15 16.54860 16.24760 15.96688 15.70508 15.46093 15.23323 15.02088 14.82284
#> 16 16.54860 16.24760 15.96688 15.70508 15.46093 15.23323 15.02088 14.82284
#> 17 23.87275 24.23712 24.57694 24.89385 25.18940 25.46504 25.72209 25.96183
#> 18 16.54860 16.24760 15.96688 15.70508 15.46093 15.23323 15.02088 14.82284
#> 19 20.21067 20.24236 20.27191 20.29947 20.32517 20.34913 20.37149 20.39233
#> 20 16.54860 16.24760 15.96688 15.70508 15.46093 15.23323 15.02088 14.82284
#> 21 23.87275 24.23712 24.57694 24.89385 25.18940 25.46504 25.72209 25.96183
#>      1.6339   1.5238   1.4211   1.3253   1.2360   1.1527   1.0750   1.0026
#> 1  20.40783 20.40783 20.40783 20.40783 20.40783 20.47890 20.85309 21.20223
#> 2  26.14006 26.14006 26.14006 26.14006 26.14006 26.15614 26.23856 26.31540
#> 3  20.40783 20.40783 20.40783 20.40783 20.40783 20.40691 20.40186 20.39713
#> 4  14.67560 14.67560 14.67560 14.67560 14.67560 14.73874 15.07327 15.38544
#> 5  20.40783 20.40783 20.40783 20.40783 20.40783 20.37727 20.21605 20.06563
#> 6  14.67560 14.67560 14.67560 14.67560 14.67560 14.72301 14.97468 15.20954
#> 7  26.14006 26.14006 26.14006 26.14006 26.14006 26.05088 25.57877 25.13821
#> 8  26.14006 26.14006 26.14006 26.14006 26.14006 26.05572 25.60910 25.19233
#> 9  20.40783 20.40783 20.40783 20.40783 20.40783 20.37969 20.23122 20.09269
#> 10 14.67560 14.67560 14.67560 14.67560 14.67560 14.66252 14.59549 14.53299
#> 11 14.67560 14.67560 14.67560 14.67560 14.67560 14.49871 13.56864 12.70090
#> 12 14.67560 14.67560 14.67560 14.67560 14.67560 14.50826 13.62855 12.80779
#> 13 26.14006 26.14006 26.14006 26.14006 26.14006 26.17066 26.32957 26.47777
#> 14 26.14006 26.14006 26.14006 26.14006 26.14006 26.21482 26.60638 26.97165
#> 15 14.67560 14.67560 14.67560 14.67560 14.67560 14.72906 15.01260 15.27719
#> 16 14.67560 14.67560 14.67560 14.67560 14.67560 14.69035 14.76991 14.84420
#> 17 26.14006 26.14006 26.14006 26.14006 26.14006 26.20272 26.53054 26.83634
#> 18 14.67560 14.67560 14.67560 14.67560 14.67560 14.77141 15.27803 15.75078
#> 19 20.40783 20.40783 20.40783 20.40783 20.40783 20.46075 20.73934 20.99926
#> 20 14.67560 14.67560 14.67560 14.67560 14.67560 14.72301 14.97468 15.20954
#> 21 26.14006 26.14006 26.14006 26.14006 26.14006 26.10049 25.88970 25.69298
#>      0.9350   0.8720   0.8132   0.7584   0.7073   0.6596   0.6152   0.5737
#> 1  21.52755 21.83123 22.03027 22.03036 22.03040 22.03043 21.97521 21.89452
#> 2  26.38711 26.45394 26.49754 26.49744 26.49739 26.49736 26.52461 26.56444
#> 3  20.39274 20.38864 20.38593 20.38592 20.38591 20.38591 20.46963 20.59199
#> 4  15.67623 15.94777 16.12593 16.12611 16.12621 16.12626 16.22131 16.36025
#> 5  19.92547 19.79463 19.70884 19.70879 19.70877 19.70875 19.79186 19.91334
#> 6  15.42829 15.63258 15.76666 15.76682 15.76691 15.76695 15.76807 15.76972
#> 7  24.72781 24.34460 24.09320 24.09296 24.09284 24.09277 24.07106 24.03930
#> 8  24.80410 24.44158 24.20374 24.20351 24.20339 24.20333 24.18171 24.15010
#> 9  19.96362 19.84312 19.76412 19.76407 19.76404 19.76403 19.70678 19.62310
#> 10 14.47467 14.42032 14.38486 14.38494 14.38498 14.38500 14.43168 14.49992
#> 11 11.89227 11.13752 10.64293 10.64280 10.64272 10.64269 10.63920 10.63410
#> 12 12.04294 11.32905 10.86126 10.86113 10.86107 10.86103 10.85775 10.85293
#> 13 26.61597 26.74488 26.82917 26.82910 26.82905 26.82903 26.85658 26.89684
#> 14 27.31212 27.62983 27.83789 27.83787 27.83786 27.83786 27.86631 27.90790
#> 15 15.52365 15.75381 15.90484 15.90501 15.90510 15.90515 16.00000 16.13865
#> 16 14.91334 14.97796 15.02049 15.02061 15.02067 15.02070 15.02115 15.02181
#> 17 27.12139 27.38738 27.56153 27.56149 27.56148 27.56147 27.58967 27.63090
#> 18 16.19119 16.60239 16.87211 16.87233 16.87245 16.87251 16.87463 16.87773
#> 19 21.24147 21.46755 21.61573 21.61579 21.61582 21.61584 21.46665 21.24859
#> 20 15.42829 15.63258 15.76666 15.76682 15.76691 15.76695 15.58086 15.30887
#> 21 25.50978 25.33866 25.22628 25.22611 25.22602 25.22597 25.20528 25.17501
#>      0.5350   0.4990   0.4653   0.4340   0.4047   0.3775   0.3520   0.3283
#> 1  21.81948 21.74932 21.68389 21.62286 21.56595 21.51288 21.46338 21.41722
#> 2  26.60148 26.63610 26.66839 26.69850 26.72659 26.75278 26.77721 26.79999
#> 3  20.70575 20.81214 20.91136 21.00390 21.09020 21.17068 21.24574 21.31574
#> 4  16.48938 16.61018 16.72284 16.82791 16.92589 17.01728 17.10250 17.18198
#> 5  20.02627 20.13189 20.23039 20.32225 20.40792 20.48781 20.56232 20.63181
#> 6  15.77122 15.77265 15.77399 15.77523 15.77639 15.77748 15.77849 15.77943
#> 7  24.00983 23.98222 23.95647 23.93246 23.91007 23.88918 23.86970 23.85154
#> 8  24.12076 24.09328 24.06765 24.04375 24.02146 24.00067 23.98128 23.96320
#> 9  19.54531 19.47256 19.40470 19.34142 19.28241 19.22737 19.17604 19.12817
#> 10 14.56334 14.62267 14.67800 14.72960 14.77772 14.82260 14.86446 14.90349
#> 11 10.62938 10.62494 10.62081 10.61695 10.61335 10.60999 10.60686 10.60394
#> 12 10.84848 10.84429 10.84039 10.83675 10.83335 10.83018 10.82723 10.82447
#> 13 26.93428 26.96928 27.00193 27.03237 27.06077 27.08725 27.11194 27.13497
#> 14 27.94656 27.98272 28.01644 28.04789 28.07722 28.10458 28.13009 28.15388
#> 15 16.26751 16.38805 16.50048 16.60533 16.70311 16.79430 16.87934 16.95866
#> 16 15.02241 15.02299 15.02352 15.02403 15.02450 15.02493 15.02534 15.02572
#> 17 27.66923 27.70507 27.73850 27.76967 27.79874 27.82586 27.85114 27.87473
#> 18 16.88057 16.88327 16.88578 16.88813 16.89032 16.89236 16.89427 16.89604
#> 19 21.04586 20.85627 20.67945 20.51455 20.36076 20.21734 20.08358 19.95884
#> 20 15.05598 14.81951 14.59896 14.39327 14.20145 14.02256 13.85572 13.70013
#> 21 25.14691 25.12060 25.09607 25.07318 25.05184 25.03194 25.01337 24.99606
#>      0.3062   0.2855   0.2663   0.2483   0.2316   0.2160   0.2014   0.1879
#> 1  21.40525 21.40515 21.40511 21.40510 21.42244 21.63517 21.62826 22.45539
#> 2  26.80590 26.80594 26.80596 26.80596 26.84618 26.36203 26.34615 25.84570
#> 3  21.33387 21.33405 21.33411 21.33414 21.30896 21.22806 21.17831 20.22440
#> 4  17.20254 17.20278 17.20287 17.20292 17.18477 17.20106 17.18992 17.71830
#> 5  20.64982 20.64999 20.65005 20.65008 20.63733 21.43642 21.32761 20.21483
#> 6  15.77965 15.77968 15.77970 15.77971 15.73444 14.56990 14.56397 14.52499
#> 7  23.84687 23.84678 23.84673 23.84671 23.79092 21.58448 21.62147 21.49468
#> 8  23.95855 23.95846 23.95842 23.95840 23.90057 25.23534 25.27339 26.42999
#> 9  19.11577 19.11564 19.11559 19.11557 19.05122 18.68774 18.76244 18.28024
#> 10 14.91359 14.91371 14.91375 14.91378 14.91075 15.35721 15.33381 15.68347
#> 11 10.60321 10.60318 10.60316 10.60315 10.65200 11.37219 11.35986 11.43271
#> 12 10.82378 10.82375 10.82373 10.82372 10.86857 11.12426 11.15524 11.38101
#> 13 27.14095 27.14099 27.14101 27.14101 27.17514 27.77178 27.78138 27.87504
#> 14 28.16004 28.16010 28.16012 28.16013 28.17573 29.39919 29.41660 29.88967
#> 15 16.97918 16.97941 16.97951 16.97955 16.96547 16.77650 16.70119 16.69409
#> 16 15.02580 15.02582 15.02584 15.02584 14.99428 13.22984 13.32927 13.72270
#> 17 27.88084 27.88089 27.88091 27.88092 27.90160 27.87952 27.88853 27.83113
#> 18 16.89647 16.89651 16.89654 16.89655 16.95430 17.58589 17.67785 18.12116
#> 19 19.92651 19.92621 19.92610 19.92605 19.91728 19.21131 19.14751 19.06985
#> 20 13.65980 13.65944 13.65930 13.65925 13.66985 14.89443 14.82798 13.99603
#> 21 24.99160 24.99153 24.99149 24.99147 25.03819 24.55768 24.58928 24.21464
#>      0.1752   0.1634   0.1524   0.1421   0.1325   0.1236   0.1153   0.1075
#> 1  22.45561 22.45578 22.45592 22.45603 22.45612 23.58485 23.58480 23.58475
#> 2  25.84559 25.84551 25.84545 25.84539 25.84535 25.68282 25.68278 25.68274
#> 3  20.22421 20.22406 20.22394 20.22384 20.22377 20.15940 20.15921 20.15906
#> 4  17.71824 17.71819 17.71816 17.71813 17.71810 17.38792 17.38793 17.38795
#> 5  20.21480 20.21478 20.21477 20.21476 20.21475 19.98400 19.98375 19.98358
#> 6  14.52484 14.52472 14.52462 14.52454 14.52448 14.62073 14.62089 14.62102
#> 7  21.49460 21.49455 21.49450 21.49447 21.49444 21.45513 21.45475 21.45445
#> 8  26.43045 26.43081 26.43110 26.43134 26.43152 25.67579 25.67626 25.67666
#> 9  18.28007 18.27993 18.27982 18.27973 18.27966 18.84139 18.84118 18.84100
#> 10 15.68353 15.68358 15.68362 15.68365 15.68368 15.78011 15.78009 15.78009
#> 11 11.43287 11.43299 11.43309 11.43317 11.43323 11.66721 11.66723 11.66725
#> 12 11.38107 11.38112 11.38116 11.38119 11.38122 11.58652 11.58665 11.58674
#> 13 27.87508 27.87512 27.87515 27.87517 27.87519 27.98799 27.98809 27.98816
#> 14 29.88979 29.88990 29.88998 29.89004 29.89010 29.97804 29.97834 29.97857
#> 15 16.69404 16.69399 16.69396 16.69393 16.69391 16.52319 16.52286 16.52262
#> 16 13.72240 13.72217 13.72198 13.72183 13.72171 14.11989 14.12015 14.12033
#> 17 27.83108 27.83104 27.83101 27.83098 27.83096 28.00357 28.00360 28.00362
#> 18 18.12095 18.12078 18.12065 18.12055 18.12046 16.69089 16.69101 16.69108
#> 19 19.07005 19.07020 19.07033 19.07043 19.07051 19.28899 19.28863 19.28836
#> 20 13.99616 13.99627 13.99636 13.99642 13.99648 13.54883 13.54897 13.54911
#> 21 24.21455 24.21449 24.21444 24.21440 24.21437 24.53273 24.53282 24.53286
#>      0.1003   0.0935   0.0872   0.0813   0.0758   0.0707   0.0660   0.0615
#> 1  23.58470 23.58465 23.58461 23.58457 23.58454 23.58451 23.58448 23.58446
#> 2  25.68270 25.68267 25.68264 25.68262 25.68260 25.68259 25.68257 25.68256
#> 3  20.15896 20.15888 20.15881 20.15877 20.15873 20.15870 20.15868 20.15866
#> 4  17.38798 17.38800 17.38803 17.38806 17.38808 17.38811 17.38813 17.38814
#> 5  19.98347 19.98340 19.98335 19.98332 19.98330 19.98329 19.98328 19.98327
#> 6  14.62112 14.62121 14.62128 14.62135 14.62141 14.62146 14.62150 14.62154
#> 7  21.45422 21.45402 21.45386 21.45371 21.45359 21.45348 21.45339 21.45331
#> 8  25.67700 25.67729 25.67753 25.67773 25.67790 25.67805 25.67817 25.67828
#> 9  18.84084 18.84070 18.84057 18.84045 18.84035 18.84026 18.84018 18.84010
#> 10 15.78009 15.78010 15.78011 15.78012 15.78013 15.78014 15.78014 15.78015
#> 11 11.66725 11.66725 11.66725 11.66724 11.66724 11.66724 11.66724 11.66723
#> 12 11.58680 11.58684 11.58686 11.58688 11.58689 11.58689 11.58690 11.58690
#> 13 27.98821 27.98825 27.98828 27.98830 27.98831 27.98832 27.98833 27.98834
#> 14 29.97875 29.97890 29.97901 29.97911 29.97919 29.97926 29.97932 29.97937
#> 15 16.52244 16.52232 16.52222 16.52215 16.52209 16.52205 16.52201 16.52198
#> 16 14.12045 14.12052 14.12058 14.12061 14.12064 14.12066 14.12067 14.12068
#> 17 28.00362 28.00362 28.00361 28.00360 28.00360 28.00359 28.00358 28.00358
#> 18 16.69113 16.69117 16.69119 16.69120 16.69121 16.69121 16.69121 16.69121
#> 19 19.28816 19.28801 19.28789 19.28780 19.28772 19.28766 19.28761 19.28756
#> 20 13.54923 13.54935 13.54946 13.54956 13.54964 13.54972 13.54979 13.54985
#> 21 24.53288 24.53288 24.53287 24.53285 24.53284 24.53283 24.53281 24.53280
#>      0.0574   0.0535   0.0499   0.0465   0.0434   0.0405   0.0377   0.0352
#> 1  23.58444 23.58443 23.58441 23.58440 23.58439 23.58438 23.53673 23.53652
#> 2  25.68256 25.68255 25.68254 25.68254 25.68254 25.68253 25.67039 25.67045
#> 3  20.15864 20.15863 20.15861 20.15860 20.15859 20.15858 20.32676 20.32687
#> 4  17.38816 17.38817 17.38818 17.38819 17.38820 17.38821 17.61040 17.61058
#> 5  19.98327 19.98327 19.98326 19.98326 19.98326 19.98325 19.97496 19.97516
#> 6  14.62157 14.62160 14.62163 14.62165 14.62167 14.62168 14.57417 14.57445
#> 7  21.45323 21.45317 21.45311 21.45306 21.45302 21.45298 21.49782 21.49720
#> 8  25.67837 25.67845 25.67851 25.67857 25.67862 25.67867 25.65903 25.65929
#> 9  18.84004 18.83998 18.83993 18.83989 18.83985 18.83982 18.71571 18.71509
#> 10 15.78016 15.78016 15.78017 15.78017 15.78018 15.78018 15.47118 15.47113
#> 11 11.66723 11.66723 11.66723 11.66723 11.66723 11.66723 11.80416 11.80420
#> 12 11.58691 11.58691 11.58692 11.58692 11.58692 11.58693 11.59691 11.59690
#> 13 27.98835 27.98836 27.98837 27.98837 27.98838 27.98838 27.97139 27.97143
#> 14 29.97941 29.97945 29.97949 29.97952 29.97954 29.97957 30.01970 30.01994
#> 15 16.52195 16.52193 16.52191 16.52189 16.52188 16.52186 16.60664 16.60663
#> 16 14.12069 14.12070 14.12071 14.12072 14.12072 14.12073 13.96519 13.96512
#> 17 28.00358 28.00357 28.00357 28.00357 28.00357 28.00357 28.03467 28.03468
#> 18 16.69122 16.69122 16.69122 16.69122 16.69122 16.69122 17.09400 17.09406
#> 19 19.28752 19.28748 19.28745 19.28743 19.28740 19.28738 19.12766 19.12738
#> 20 13.54990 13.54995 13.54999 13.55002 13.55005 13.55008 13.41048 13.41096
#> 21 24.53279 24.53278 24.53278 24.53277 24.53277 24.53276 24.43205 24.43199
#>      0.0328   0.0306   0.0286   0.0266   0.0248   0.0232   0.0216   0.0201
#> 1  23.53634 23.53619 23.53607 23.53596 23.53587 23.53580 23.70357 23.70370
#> 2  25.67050 25.67055 25.67058 25.67062 25.67065 25.67067 25.56714 25.56704
#> 3  20.32693 20.32697 20.32698 20.32697 20.32693 20.32689 20.41270 20.41288
#> 4  17.61072 17.61082 17.61088 17.61092 17.61093 17.61093 17.62141 17.62152
#> 5  19.97532 19.97546 19.97557 19.97566 19.97573 19.97579 19.95137 19.95138
#> 6  14.57470 14.57493 14.57514 14.57532 14.57549 14.57564 14.64771 14.64784
#> 7  21.49664 21.49613 21.49568 21.49528 21.49491 21.49459 21.46556 21.46543
#> 8  25.65953 25.65976 25.65998 25.66019 25.66038 25.66056 25.64745 25.64745
#> 9  18.71456 18.71409 18.71369 18.71335 18.71306 18.71282 18.81037 18.81034
#> 10 15.47112 15.47114 15.47118 15.47125 15.47133 15.47143 15.28161 15.28130
#> 11 11.80422 11.80422 11.80422 11.80420 11.80418 11.80415 11.88378 11.88390
#> 12 11.59688 11.59688 11.59687 11.59687 11.59688 11.59688 11.58818 11.58817
#> 13 27.97147 27.97152 27.97156 27.97160 27.97164 27.97167 27.93910 27.93906
#> 14 30.02015 30.02035 30.02052 30.02068 30.02082 30.02094 30.07559 30.07570
#> 15 16.60659 16.60655 16.60649 16.60642 16.60635 16.60627 16.54085 16.54080
#> 16 13.96507 13.96505 13.96506 13.96508 13.96512 13.96517 13.95040 13.95034
#> 17 28.03468 28.03468 28.03468 28.03468 28.03468 28.03467 28.04084 28.04086
#> 18 17.09407 17.09405 17.09399 17.09391 17.09381 17.09370 16.94367 16.94363
#> 19 19.12715 19.12695 19.12678 19.12664 19.12652 19.12643 19.14216 19.14208
#> 20 13.41140 13.41180 13.41216 13.41250 13.41280 13.41307 13.49452 13.49466
#> 21 24.43194 24.43192 24.43190 24.43190 24.43191 24.43193 24.39202 24.39194
#>      0.0188   0.0175   0.0163   0.0152   0.0142   0.0133   0.0124   0.0115
#> 1  23.70383 23.70395 23.70407 23.70418 23.74538 23.74549 23.74558 23.74568
#> 2  25.56694 25.56684 25.56674 25.56665 25.48251 25.48243 25.48236 25.48229
#> 3  20.41304 20.41321 20.41336 20.41351 20.42539 20.42558 20.42575 20.42592
#> 4  17.62163 17.62173 17.62183 17.62193 17.62731 17.62742 17.62752 17.62762
#> 5  19.95139 19.95140 19.95141 19.95142 20.00920 20.00921 20.00922 20.00923
#> 6  14.64796 14.64808 14.64819 14.64830 14.54284 14.54296 14.54309 14.54321
#> 7  21.46531 21.46519 21.46507 21.46496 21.31149 21.31140 21.31132 21.31124
#> 8  25.64744 25.64744 25.64744 25.64744 25.61460 25.61457 25.61455 25.61453
#> 9  18.81031 18.81028 18.81026 18.81023 18.98269 18.98265 18.98259 18.98254
#> 10 15.28099 15.28070 15.28042 15.28015 15.39190 15.39157 15.39124 15.39092
#> 11 11.88403 11.88415 11.88426 11.88437 11.86292 11.86305 11.86317 11.86330
#> 12 11.58815 11.58814 11.58812 11.58811 11.56824 11.56823 11.56821 11.56820
#> 13 27.93902 27.93898 27.93895 27.93892 27.97180 27.97176 27.97171 27.97167
#> 14 30.07580 30.07590 30.07599 30.07609 30.13823 30.13831 30.13839 30.13847
#> 15 16.54076 16.54072 16.54068 16.54064 16.56825 16.56821 16.56817 16.56813
#> 16 13.95027 13.95021 13.95015 13.95010 13.89641 13.89633 13.89627 13.89621
#> 17 28.04087 28.04089 28.04091 28.04092 28.03815 28.03817 28.03819 28.03821
#> 18 16.94360 16.94356 16.94353 16.94349 17.01646 17.01644 17.01643 17.01641
#> 19 19.14200 19.14193 19.14186 19.14179 19.06953 19.06946 19.06940 19.06934
#> 20 13.49479 13.49492 13.49504 13.49516 13.50875 13.50888 13.50900 13.50912
#> 21 24.39186 24.39178 24.39170 24.39163 24.32797 24.32790 24.32783 24.32777
#>      0.0108   0.0100   0.0093   0.0087   0.0081   0.0076   0.0071   0.0066
#> 1  23.74577 23.74586 23.74595 23.74603 23.74612 23.74619 23.74627 23.74634
#> 2  25.48222 25.48216 25.48210 25.48204 25.48199 25.48194 25.48189 25.48184
#> 3  20.42609 20.42625 20.42640 20.42655 20.42669 20.42682 20.42695 20.42708
#> 4  17.62772 17.62782 17.62791 17.62800 17.62808 17.62817 17.62825 17.62832
#> 5  20.00924 20.00924 20.00924 20.00924 20.00924 20.00924 20.00923 20.00923
#> 6  14.54333 14.54345 14.54357 14.54369 14.54381 14.54393 14.54404 14.54416
#> 7  21.31117 21.31110 21.31104 21.31098 21.31093 21.31089 21.31084 21.31080
#> 8  25.61451 25.61450 25.61448 25.61447 25.61446 25.61445 25.61445 25.61444
#> 9  18.98248 18.98242 18.98236 18.98230 18.98223 18.98217 18.98210 18.98203
#> 10 15.39061 15.39031 15.39001 15.38972 15.38944 15.38917 15.38890 15.38865
#> 11 11.86342 11.86354 11.86365 11.86376 11.86387 11.86397 11.86407 11.86417
#> 12 11.56819 11.56818 11.56817 11.56816 11.56815 11.56815 11.56814 11.56813
#> 13 27.97163 27.97159 27.97155 27.97151 27.97147 27.97143 27.97140 27.97136
#> 14 30.13854 30.13861 30.13868 30.13874 30.13880 30.13886 30.13891 30.13896
#> 15 16.56809 16.56805 16.56802 16.56798 16.56794 16.56791 16.56787 16.56784
#> 16 13.89615 13.89609 13.89604 13.89600 13.89595 13.89591 13.89587 13.89584
#> 17 28.03823 28.03825 28.03826 28.03828 28.03830 28.03831 28.03833 28.03834
#> 18 17.01639 17.01636 17.01634 17.01631 17.01628 17.01625 17.01622 17.01618
#> 19 19.06928 19.06923 19.06918 19.06913 19.06909 19.06905 19.06901 19.06898
#> 20 13.50923 13.50934 13.50945 13.50955 13.50965 13.50975 13.50984 13.50993
#> 21 24.32771 24.32765 24.32760 24.32755 24.32750 24.32746 24.32741 24.32737
#>      0.0062   0.0057   0.0054   0.0050
#> 1  23.74641 23.74648 23.74655 23.74661
#> 2  25.48180 25.48176 25.48173 25.48169
#> 3  20.42720 20.42731 20.42743 20.42753
#> 4  17.62840 17.62847 17.62854 17.62860
#> 5  20.00922 20.00922 20.00921 20.00920
#> 6  14.54427 14.54438 14.54449 14.54460
#> 7  21.31077 21.31074 21.31071 21.31069
#> 8  25.61444 25.61444 25.61444 25.61444
#> 9  18.98196 18.98189 18.98182 18.98175
#> 10 15.38840 15.38815 15.38792 15.38769
#> 11 11.86426 11.86435 11.86444 11.86453
#> 12 11.56813 11.56812 11.56812 11.56811
#> 13 27.97132 27.97129 27.97125 27.97122
#> 14 30.13901 30.13906 30.13910 30.13914
#> 15 16.56780 16.56777 16.56773 16.56770
#> 16 13.89580 13.89577 13.89574 13.89572
#> 17 28.03836 28.03837 28.03839 28.03840
#> 18 17.01614 17.01611 17.01607 17.01603
#> 19 19.06894 19.06892 19.06889 19.06886
#> 20 13.51002 13.51010 13.51018 13.51026
#> 21 24.32734 24.32730 24.32727 24.32724
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 22.8 21.4 18.7 18.1 14.3 24.4 22.8 17.8 16.4 10.4 14.7 32.4 33.9 15.5
#> [16] 13.3 27.3 15.8 19.7 15.0 21.4
#> 
#> $X
#>               am        carb        cyl        disp        drat       gear
#>  [1,]  1.1547005  0.60070445 -0.1097643 -0.61631727  0.68189524  0.4082483
#>  [2,]  1.1547005 -1.04470340 -1.2622890 -1.04865391  0.57597956  0.4082483
#>  [3,] -0.8660254 -1.04470340 -0.1097643  0.19847103 -1.05512192 -1.0206207
#>  [4,] -0.8660254 -0.49623411  1.0427605  1.04651599 -0.90683997 -1.0206207
#>  [5,] -0.8660254 -1.04470340 -0.1097643 -0.07589646 -1.73298228 -1.0206207
#>  [6,] -0.8660254  0.60070445  1.0427605  1.04651599 -0.77974115 -1.0206207
#>  [7,] -0.8660254 -0.49623411 -1.2622890 -0.72689568  0.23704938  0.4082483
#>  [8,] -0.8660254 -0.49623411 -1.2622890 -0.77594926  0.72426151  0.4082483
#>  [9,] -0.8660254  0.60070445 -0.1097643 -0.55312960  0.72426151  0.4082483
#> [10,] -0.8660254  0.05223517  1.0427605  0.34646319 -1.07630506 -1.0206207
#> [11,] -0.8660254  0.60070445  1.0427605  1.87793261 -1.22458701 -1.0206207
#> [12,] -0.8660254  0.60070445  1.0427605  1.71164929 -0.73737488 -1.0206207
#> [13,]  1.1547005 -1.04470340 -1.2622890 -1.29225899  1.06319169  0.4082483
#> [14,]  1.1547005 -1.04470340 -1.2622890 -1.35544665  1.35975559  0.4082483
#> [15,] -0.8660254 -0.49623411  1.0427605  0.69732100 -1.73298228 -1.0206207
#> [16,] -0.8660254  0.60070445  1.0427605  0.96337432  0.32178192 -1.0206207
#> [17,]  1.1547005 -1.04470340 -1.2622890 -1.28976474  1.06319169  0.4082483
#> [18,]  1.1547005  0.60070445  1.0427605  0.97168849  1.35975559  1.8371173
#> [19,]  1.1547005  1.69764302 -0.1097643 -0.74102976  0.08876743  1.8371173
#> [20,]  1.1547005  2.79458158  1.0427605  0.55598018 -0.08069766  1.8371173
#> [21,]  1.1547005 -0.49623411 -1.2622890 -0.94056975  1.12674110  0.4082483
#>               hp        qsec         vs          wt
#>  [1,] -0.5700518 -0.69409287 -0.9534626 -0.73098158
#>  [2,] -0.7941801  0.34635511  1.0488088 -1.06362995
#>  [3,] -0.5700518  0.74801642  1.0488088 -0.07122899
#>  [4,]  0.2869093 -0.42309247 -0.9534626  0.17825728
#>  [5,] -0.6359719  1.12548127  1.0488088  0.20043384
#>  [6,]  1.2097905 -0.99412903 -0.9534626  0.32240491
#>  [7,] -1.2028846  1.01901682  1.0488088 -0.09894969
#>  [8,] -0.7678121  2.42241177  1.0488088 -0.14330280
#>  [9,] -0.3986596  0.48669460  1.0488088  0.17825728
#> [10,]  0.3528294 -0.23919934 -0.9534626  0.87681885
#> [11,]  0.8142700 -0.03594903 -0.9534626  2.37817181
#> [12,]  1.0120303 -0.22952075 -0.9534626  2.29057441
#> [13,] -1.1501486  0.76253430  1.0488088 -1.19668930
#> [14,] -1.1633326  0.97062390  1.0488088 -1.60141148
#> [15,] -0.0426911 -0.49568186 -0.9534626  0.26696351
#> [16,]  1.2097905 -1.20221863 -0.9534626  0.62178844
#> [17,] -1.1501486  0.48669460  1.0488088 -1.49052869
#> [18,]  1.4602869 -1.64259428 -0.9534626 -0.12112625
#> [19,]  0.2869093 -1.15866499 -0.9534626 -0.56465740
#> [20,]  2.3963521 -1.59420136 -0.9534626  0.32240491
#> [21,] -0.5832358  0.34151582  1.0488088 -0.55356912
#> attr(,"center")
#>  [1]   0.4285714   2.9047619   6.1904762 234.1285714   3.5780952   3.7142857
#>  [7] 153.2380952  17.8942857   0.4761905   3.2792381
#> attr(,"scale")
#>  [1]   0.4948717   1.8232562   1.7353207 120.2766420   0.4720736   0.6998542
#>  [7]  75.8494152   2.0664176   0.4994328   0.9018532
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 30
#> 
#> $lambda.min
#> [1] 0.6596197
#> 
#> $null.dev
#> [1] 34.45379
#> 
#> $Bias
#> [1] 4.857896
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 4.731702