Cross-Validated MCP- or SCAD-penalized Regression Learner

MCP- or SCAD-penalized regression models with k-fold cross validation for choosing the regularization parameter lambda. Calls ncvreg::cv.ncvreg() from package ncvreg with family set to "gaussian" (default). Poisson family is also supported. Covariates are internally standardized by ncvreg prior to model fitting.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.cv_ncvreg")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “integer”, “numeric”

Required Packages: mlr3, mlr3extralearners, ncvreg

Parameters

Id	Type	Default	Levels	Range

family	character	gaussian	gaussian, poisson	-

penalty	character	MCP	MCP, SCAD, lasso	-

gamma	numeric	-		\([0, \infty)\)

alpha	numeric	1		\([0, 1]\)

lambda.min	numeric	-		\([0, \infty)\)

nlambda	integer	100		\([1, \infty)\)

lambda	untyped	-		-

eps	numeric	1e-04		\((-\infty, \infty)\)

max.iter	integer	10000		\([1, \infty)\)

convex	logical	TRUE	TRUE, FALSE	-

dfmax	integer	-		\([0, \infty)\)

penalty.factor	untyped	-		-

warn	logical	TRUE	TRUE, FALSE	-

returnX	logical	TRUE	TRUE, FALSE	-

cluster	untyped	-		-

nfolds	integer	10		\((-\infty, \infty)\)

fold	untyped	-		-

returnY	logical	FALSE	TRUE, FALSE	-

trace	logical	FALSE	TRUE, FALSE	-

pred_lambda	numeric	-		\((-\infty, \infty)\)

References

Breheny P, Huang J (2011). “Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection.” Annals of Applied Statistics, 5(1), 232–253. ISSN 1932-6157, doi:10.1214/10-AOAS388 .

See also

Author

bblodfon

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNCVreg

Methods

Public methods

LearnerRegrNCVreg$new()

LearnerRegrNCVreg$selected_features()

LearnerRegrNCVreg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNCVreg$new()

Method selected_features()

Returns the set of selected features which have non-zero coefficients. Calls the internal coef.cv.ncvreg() function.

Usage

LearnerRegrNCVreg$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the lambda with the minimum CV error.

Returns

(character()) vector of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNCVreg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("regr.cv_ncvreg")
print(learner)
#> 
#> ── <LearnerRegrNCVreg> (regr.cv_ncvreg): MCP- or SCAD-penalized Regression Model
#> • Model: -
#> • Parameters: list()
#> • Packages: mlr3, mlr3extralearners, and ncvreg
#> • Predict Types: [response]
#> • Feature Types: integer and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: selected_features
#> • Other settings: use_weights = 'error'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $cve
#>   [1] 38.65481 36.99869 34.44382 32.43981 30.90121 29.57329 26.97696 24.84783
#>   [9] 23.11634 21.72261 20.61507 19.74938 19.08748 18.59669 18.24898 17.95098
#>  [17] 17.44121 16.88816 16.41180 15.98158 15.59256 15.39192 15.45915 15.44330
#>  [25] 15.63371 16.05391 16.62573 17.79281 18.50582 18.37250 18.01638 17.91078
#>  [33] 17.28548 16.83718 18.16527 16.55419 15.72286 15.72262 15.49007 15.27319
#>  [41] 15.17023 15.18025 15.20160 15.51045 15.20528 15.00534 15.00531 15.11229
#>  [49] 16.55615 16.55649 15.37643 15.93256 15.91855 16.67238 16.53392 16.43944
#>  [57] 16.53504 16.76708 16.32680 16.28028 16.35517 16.48826 16.48783 16.13010
#>  [65] 15.94881 15.57359 15.58573 16.34308 15.70345 15.16501 15.16626 14.73896
#>  [73] 14.87806 14.87923 15.11032 15.10948 15.02789 15.02707 15.02629 14.03641
#>  [81] 14.05308 14.05247 14.17896 14.16835 14.16821 14.74625 14.74735 14.74842
#>  [89] 14.74945 14.81755 14.81851 14.81943 14.82032 14.82118 14.82203 14.81626
#>  [97] 14.81690 14.81750 14.78214 14.78247
#> 
#> $cvse
#>   [1] 14.005981 14.051190 13.951236 13.888377 13.846166 13.694379 12.480443
#>   [8] 11.391335 10.413828  9.536963  8.751686  8.050533  7.427340  6.876974
#>  [15]  6.395076  5.978443  5.623947  5.300230  5.008397  4.748865  4.518057
#>  [22]  4.382210  4.382949  4.357764  4.313172  4.268567  4.246406  4.271705
#>  [29]  4.327700  4.256718  4.211507  4.203578  4.197245  4.051745  3.811232
#>  [36]  3.565008  3.455556  3.455472  3.455699  3.484153  3.364459  3.365724
#>  [43]  3.300415  3.340232  3.252809  3.223340  3.223377  3.272672  3.923407
#>  [50]  3.923620  3.817532  3.784885  3.734413  3.821895  3.836707  3.809963
#>  [57]  3.795859  3.660902  3.555456  3.564008  3.568426  3.524307  3.509747
#>  [64]  3.478631  3.814003  3.840488  3.847203  3.900862  3.919779  3.835987
#>  [71]  3.836462  3.596733  3.573969  3.574406  3.698918  3.698221  3.657825
#>  [78]  3.657155  3.656516  3.297029  3.298800  3.298338  3.309600  3.309145
#>  [85]  3.308796  3.704282  3.704826  3.705357  3.705875  3.715232  3.715739
#>  [92]  3.716233  3.716714  3.717183  3.717637  3.718287  3.718726  3.719152
#>  [99]  3.717645  3.718032
#> 
#> $fold
#>  [1]  6  3  1  9  9  6  7  1  3  4  4  2  8  8  5  2  1  7  5 10 10
#> 
#> $lambda
#>   [1] 5.079267981 4.736942319 4.417688261 4.119950858 3.842279959 3.583323149
#>   [7] 3.341819162 3.116591735 2.906543883 2.710652553 2.527963643 2.357587355
#>  [13] 2.198693857 2.050509250 1.912311790 1.783428375 1.663231272 1.551135051
#>  [19] 1.446593740 1.349098163 1.258173462 1.173376782 1.094295114 1.020543286
#>  [25] 0.951762084 0.887616505 0.827794123 0.772003570 0.719973113 0.671449335
#>  [31] 0.626195897 0.583992389 0.544633257 0.507926798 0.473694232 0.441768826
#>  [37] 0.411995086 0.384227996 0.358332315 0.334181916 0.311659174 0.290654388
#>  [43] 0.271065255 0.252796364 0.235758735 0.219869386 0.205050925 0.191231179
#>  [49] 0.178342838 0.166323127 0.155113505 0.144659374 0.134909816 0.125817346
#>  [55] 0.117337678 0.109429511 0.102054329 0.095176208 0.088761650 0.082779412
#>  [61] 0.077200357 0.071997311 0.067144933 0.062619590 0.058399239 0.054463326
#>  [67] 0.050792680 0.047369423 0.044176883 0.041199509 0.038422800 0.035833231
#>  [73] 0.033418192 0.031165917 0.029065439 0.027106526 0.025279636 0.023575874
#>  [79] 0.021986939 0.020505093 0.019123118 0.017834284 0.016632313 0.015511351
#>  [85] 0.014465937 0.013490982 0.012581735 0.011733768 0.010942951 0.010205433
#>  [91] 0.009517621 0.008876165 0.008277941 0.007720036 0.007199731 0.006714493
#>  [97] 0.006261959 0.005839924 0.005446333 0.005079268
#> 
#> $fit
#> $beta
#>              5.0793       4.7369       4.4177      4.1200      3.8423
#> (Intercept) 18.2619 19.357559207 20.379370209 21.33231457 22.22103367
#> am           0.0000  0.000000000  0.000000000  0.00000000  0.00000000
#> carb         0.0000  0.000000000  0.000000000  0.00000000  0.00000000
#> cyl          0.0000  0.000000000  0.000000000  0.00000000  0.00000000
#> disp         0.0000 -0.004140125 -0.008001219 -0.01160209 -0.01496027
#> drat         0.0000  0.000000000  0.000000000  0.00000000  0.00000000
#> gear         0.0000  0.000000000  0.000000000  0.00000000  0.00000000
#> hp           0.0000  0.000000000  0.000000000  0.00000000  0.00000000
#> qsec         0.0000  0.000000000  0.000000000  0.00000000  0.00000000
#> vs           0.0000  0.000000000  0.000000000  0.00000000  0.00000000
#> wt           0.0000  0.000000000  0.000000000  0.00000000  0.00000000
#>                  3.5833     3.3418      3.1166      2.9065     2.7107
#> (Intercept) 23.04985608 23.8228186 24.54368609 25.21596950 25.8429433
#> am           0.00000000  0.0000000  0.00000000  0.00000000  0.0000000
#> carb         0.00000000  0.0000000  0.00000000  0.00000000  0.0000000
#> cyl          0.00000000  0.0000000  0.00000000  0.00000000  0.0000000
#> disp        -0.01809212 -0.0210129 -0.02373683 -0.02627717 -0.0286463
#> drat         0.00000000  0.0000000  0.00000000  0.00000000  0.0000000
#> gear         0.00000000  0.0000000  0.00000000  0.00000000  0.0000000
#> hp           0.00000000  0.0000000  0.00000000  0.00000000  0.0000000
#> qsec         0.00000000  0.0000000  0.00000000  0.00000000  0.0000000
#> vs           0.00000000  0.0000000  0.00000000  0.00000000  0.0000000
#> wt           0.00000000  0.0000000  0.00000000  0.00000000  0.0000000
#>                  2.5280      2.3576      2.1987      2.0505      1.9123
#> (Intercept) 26.42766108 26.97297088 27.48152862 27.95581127 28.39812886
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.03085576 -0.03291631 -0.03483798 -0.03663015 -0.03830152
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  1.7834      1.6632      1.5511      1.4466      1.3491
#> (Intercept) 28.81063573 29.09977672 29.09977672 29.09977672 29.09977672
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.03986025 -0.04095282 -0.04095282 -0.04095282 -0.04095282
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  1.2582      1.1734      1.0943      1.0205      0.9518
#> (Intercept) 29.09977672 29.09977672 29.09977672 29.09977672 29.09977672
#> am           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> carb         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.04095282 -0.04095282 -0.04095282 -0.04095282 -0.04095282
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.8876      0.8278      0.7720      0.7200      0.6714
#> (Intercept) 29.09977672 29.07672066 28.76393872 28.47185814 27.23615144
#> am           0.00000000  0.02544657  0.37065854  0.69302278  2.69704939
#> carb         0.00000000  0.00000000  0.00000000  0.00000000 -0.52335690
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.04095282 -0.04089775 -0.04015067 -0.03945302 -0.03146926
#> drat         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.6262      0.5840      0.5446     0.5079      0.4737
#> (Intercept) 26.37306847 26.33171599 26.32630028 26.3259947 26.32580649
#> am           4.15102658  4.32715304  4.35015282  4.3506488  4.35095438
#> carb        -0.93325326 -1.03992480 -1.05384059 -1.0539704 -1.05405034
#> cyl          0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> disp        -0.02546647 -0.02434202 -0.02419528 -0.0241933 -0.02419208
#> drat         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> gear         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> vs           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#>                  0.4418      0.4120     0.3842      0.3583      0.3342
#> (Intercept) 26.32569052 26.32561909 18.5129747 18.51287837 18.51303695
#> am           4.35114262  4.35125859  3.0521190  3.05280232  3.05330391
#> carb        -1.05409960 -1.05412994 -1.1432499 -1.14342845 -1.14354126
#> cyl          0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> disp        -0.02419133 -0.02419087 -0.0202244 -0.02022169 -0.02022008
#> drat         0.00000000  0.00000000  2.1355325  2.13544007  2.13532054
#> gear         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> hp           0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> qsec         0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> vs           0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#>                  0.3117      0.2907      0.2711     0.2528      0.2358
#> (Intercept) 18.51328663 18.51354319 18.51376863 18.5139504  8.09555037
#> am           3.05365873  3.05390172  3.05406313  3.0541671  3.57316914
#> carb        -1.14361013 -1.14365049 -1.14367293 -1.1436845 -0.95645786
#> cyl          0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> disp        -0.02021918 -0.02021871 -0.02021848 -0.0202184 -0.01681564
#> drat         2.13520498  2.13510655  2.13502875  2.1349704  2.32001624
#> gear         0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> hp           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> qsec         0.00000000  0.00000000  0.00000000  0.0000000  0.46376942
#> vs           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.0000000  0.00000000
#>                  0.2199      0.2051      0.1912      0.1783      0.1663
#> (Intercept) 13.58555984 13.58584053 13.58601053 13.58607874 13.58606283
#> am           5.85101785  5.85190404  5.85263774  5.85324675  5.85375341
#> carb        -0.60056267 -0.60043481 -0.60031934 -0.60021680 -0.60012704
#> cyl          0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> disp        -0.01822459 -0.01822445 -0.01822438 -0.01822433 -0.01822429
#> drat         2.67570504  2.67581590  2.67592743  2.67603582  2.67613825
#> gear        -2.33243260 -2.33309103 -2.33365695 -2.33414033 -2.33455087
#> hp           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> qsec         0.46026511  0.46031698  0.46036272  0.46040346  0.46043988
#> vs           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>                  0.1551      0.1447     0.1349      0.1258      0.1173
#> (Intercept) 13.58598321 13.58585960 13.5857092 13.58554584 13.58538023
#> am           5.85417574  5.85452834  5.8548231  5.85506973  5.85527622
#> carb        -0.60004938 -0.59998289 -0.5999264 -0.59987887 -0.59983902
#> cyl          0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> disp        -0.01822424 -0.01822418 -0.0182241 -0.01822402 -0.01822394
#> drat         2.67623284  2.67631850  2.6763948  2.67646189  2.67652011
#> gear        -2.33489782 -2.33518975 -2.3354345 -2.33563893 -2.33580935
#> hp           0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> qsec         0.46047242  0.46050138  0.4605270  0.46054954  0.46056923
#> vs           0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#> wt           0.00000000  0.00000000  0.0000000  0.00000000  0.00000000
#>                    0.1094        0.1021        0.0952        0.0888
#> (Intercept) -22.804723811 -22.797836180 -22.791314878 -22.785148640
#> am            8.035815972   8.035375058   8.034953731   8.034551719
#> carb          0.000000000   0.000000000   0.000000000   0.000000000
#> cyl           1.545041277   1.544729506   1.544437186   1.544163475
#> disp         -0.006288002  -0.006289275  -0.006290621  -0.006292023
#> drat          3.718609558   3.718340673   3.718091773   3.717861760
#> gear         -3.285711993  -3.285652185  -3.285596874  -3.285545808
#> hp            0.000000000   0.000000000   0.000000000   0.000000000
#> qsec          2.083482290   2.083234368   2.082997584   2.082771769
#> vs            0.000000000   0.000000000   0.000000000   0.000000000
#> wt           -2.488916491  -2.488686584  -2.488457440  -2.488230012
#>                   0.0828        0.0772        0.0720       0.0671        0.0626
#> (Intercept) -22.77932585 -22.773834618 -22.768662893 -22.76379851 -30.204100886
#> am            8.03416869   8.033804268   8.033458028   8.03312952   8.705278505
#> carb          0.00000000   0.000000000   0.000000000   0.00000000   0.000000000
#> cyl           1.54390754   1.543668561   1.543445730   1.54323826   1.714969975
#> disp         -0.00629347  -0.006294951  -0.006296453  -0.00629797  -0.007282920
#> drat          3.71764957   3.717454183   3.717274608   3.71710990   4.012665777
#> gear         -3.28549875  -3.285455457  -3.285415715  -3.28537931  -3.805286574
#> hp            0.00000000   0.000000000   0.000000000   0.00000000   0.009890125
#> qsec          2.08255672   2.082352225   2.082158023   2.08197385   2.447539873
#> vs            0.00000000   0.000000000   0.000000000   0.00000000   0.000000000
#> wt           -2.48800514  -2.487783561  -2.487565909  -2.48735273  -2.767338996
#>                    0.0584        0.0545        0.0508       0.0474
#> (Intercept) -30.211985140 -30.219516990 -30.226709814 -30.23357667
#> am            8.705975448   8.706646302   8.707291874   8.70791295
#> carb          0.000000000   0.000000000   0.000000000   0.00000000
#> cyl           1.715235563   1.715487408   1.715726205   1.71595262
#> disp         -0.007281887  -0.007280880  -0.007279904  -0.00727896
#> drat          4.012961526   4.013243101   4.013511236   4.01376663
#> gear         -3.805618876  -3.805941110  -3.806253513  -3.80655631
#> hp            0.009895008   0.009899752   0.009904355   0.00990882
#> qsec          2.447883757   2.448213625   2.448529893   2.44883298
#> vs            0.000000000   0.000000000   0.000000000   0.00000000
#> wt           -2.767670991  -2.767991266  -2.768299846  -2.76859680
#>                    0.0442        0.0412        0.0384        0.0358
#> (Intercept) -30.240130302 -29.577190328 -29.572539217 -29.568106398
#> am            8.708510286   8.636374439   8.635743956   8.635141458
#> carb          0.000000000  -0.202971986  -0.203049974  -0.203121664
#> cyl           1.716167277   1.717437844   1.717384959   1.717333003
#> disp         -0.007278049  -0.009817931  -0.009820947  -0.009823738
#> drat          4.014009928   4.093403128   4.093331534   4.093259344
#> gear         -3.806849711  -3.669089836  -3.668657529  -3.668244708
#> hp            0.009913148   0.012469803   0.012464705   0.012459808
#> qsec          2.449123292   2.352262593   2.351966755   2.351686406
#> vs            0.000000000   0.000000000   0.000000000   0.000000000
#> wt           -2.768882223  -2.435246159  -2.434742992  -2.434272091
#>                    0.0334        0.0312        0.0291       0.0271       0.0253
#> (Intercept) -29.563882393 -29.559858088 -29.556024719 -27.13634956 -27.13820778
#> am            8.634565934   8.634016389   8.633491851   9.37007916   9.37103029
#> carb         -0.203187448  -0.203247697  -0.203302766  -0.14607071  -0.14601651
#> cyl           1.717281974   1.717231872   1.717182697   2.02655799   2.02688753
#> disp         -0.009826317  -0.009828698  -0.009830893  -0.01017650  -0.01017634
#> drat          4.093186763   4.093113980   4.093041174   4.05242675   4.05254530
#> gear         -3.667850687  -3.667474792  -3.667116359  -4.21359088  -4.21421332
#> hp            0.012455108   0.012450599   0.012446275   0.01022672   0.01022750
#> qsec          2.351420825   2.351169319   2.350931223   2.16774268   2.16778674
#> vs            0.000000000   0.000000000   0.000000000   1.40779575   1.40867408
#> wt           -2.433831701  -2.433420148  -2.433035829  -2.33953112  -2.33962838
#>                   0.0236       0.0220       0.0205       0.0191       0.0178
#> (Intercept) -27.14002455 -27.14180032 -27.14353562 -27.14523095 -27.14688686
#> am            9.37195373   9.37285023   9.37372057   9.37456547   9.37538565
#> carb         -0.14596410  -0.14591344  -0.14586447  -0.14581712  -0.14577135
#> cyl           2.02720747   2.02751809   2.02781965   2.02811240   2.02839660
#> disp         -0.01017619  -0.01017605  -0.01017591  -0.01017578  -0.01017565
#> drat          4.05266101   4.05277393   4.05288412   4.05299161   4.05309647
#> gear         -4.21481744  -4.21540378  -4.21597283  -4.21652508  -4.21706103
#> hp            0.01022826   0.01022902   0.01022977   0.01023050   0.01023122
#> qsec          2.16783008   2.16787269   2.16791456   2.16795567   2.16799603
#> vs            1.40952530   1.41035027   1.41114979   1.41192467   1.41267566
#> wt           -2.33972288  -2.33981469  -2.33990386  -2.33999045  -2.34007453
#>                   0.0166       0.0155       0.0145       0.0135       0.0126
#> (Intercept) -27.14850391 -27.15008266 -27.15162370 -27.15312763 -27.15459504
#> am            9.37618180   9.37695461   9.37770473   9.37843281   9.37913948
#> carb         -0.14572710  -0.14568433  -0.14564298  -0.14560300  -0.14556436
#> cyl           2.02867249   2.02894032   2.02920030   2.02945266   2.02969763
#> disp         -0.01017553  -0.01017542  -0.01017531  -0.01017521  -0.01017512
#> drat          4.05319875   4.05329849   4.05339575   4.05349057   4.05358300
#> gear         -4.21758112  -4.21808583  -4.21857559  -4.21905083  -4.21951198
#> hp            0.01023194   0.01023264   0.01023332   0.01023400   0.01023466
#> qsec          2.16803561   2.16807443   2.16811248   2.16814976   2.16818627
#> vs            1.41340353   1.41410898   1.41479272   1.41545543   1.41609775
#> wt           -2.34015616  -2.34023539  -2.34031228  -2.34038689  -2.34045928
#>                   0.0117       0.0109       0.0102       0.0095       0.0089
#> (Intercept) -27.15602655 -27.15742278 -27.15878434 -27.16011186 -27.16140597
#> am            9.37982535   9.38049100   9.38113703   9.38176399   9.38237244
#> carb         -0.14552701  -0.14549090  -0.14545599  -0.14542225  -0.14538963
#> cyl           2.02993542   2.03016623   2.03039026   2.03060771   2.03081877
#> disp         -0.01017503  -0.01017494  -0.01017486  -0.01017478  -0.01017471
#> drat          4.05367310   4.05376091   4.05384647   4.05392985   4.05401108
#> gear         -4.21995943  -4.22039359  -4.22081484  -4.22122355  -4.22162010
#> hp            0.01023532   0.01023596   0.01023658   0.01023720   0.01023780
#> qsec          2.16822200   2.16825697   2.16829118   2.16832462   2.16835732
#> vs            1.41672034   1.41732379   1.41790871   1.41847567   1.41902524
#> wt           -2.34052951  -2.34059763  -2.34066369  -2.34072776  -2.34078987
#>                   0.0083       0.0077       0.0072       0.0067       0.0063
#> (Intercept) -27.16266731 -27.16389650 -27.16509417 -27.16626095 -27.16739747
#> am            9.38296289   9.38353588   9.38409191   9.38463146   9.38515502
#> carb         -0.14535809  -0.14532761  -0.14529814  -0.14526964  -0.14524210
#> cyl           2.03102362   2.03122244   2.03141541   2.03160270   2.03178447
#> disp         -0.01017464  -0.01017458  -0.01017451  -0.01017446  -0.01017440
#> drat          4.05409021   4.05416728   4.05424235   4.05431545   4.05438663
#> gear         -4.22200484  -4.22237811  -4.22274024  -4.22309157  -4.22343241
#> hp            0.01023839   0.01023897   0.01023954   0.01024009   0.01024063
#> qsec          2.16838927   2.16842048   2.16845095   2.16848071   2.16850975
#> vs            1.41955796   1.42007434   1.42057491   1.42106015   1.42153053
#> wt           -2.34085010  -2.34090848  -2.34096507  -2.34101992  -2.34107308
#>                   0.0058       0.0054       0.0051
#> (Intercept) -27.16850435 -27.16958221 -27.17063167
#> am            9.38566305   9.38615600   9.38663432
#> carb         -0.14521547  -0.14518972  -0.14516483
#> cyl           2.03196088   2.03213209   2.03229825
#> disp         -0.01017435  -0.01017431  -0.01017426
#> drat          4.05445593   4.05452340   4.05458908
#> gear         -4.22376307  -4.22408385  -4.22439504
#> hp            0.01024116   0.01024168   0.01024219
#> qsec          2.16853808   2.16856572   2.16859268
#> vs            1.42198653   1.42242858   1.42285713
#> wt           -2.34112459  -2.34117450  -2.34122286
#> 
#> $iter
#>   [1]   0   7   3   2   2   2   2   2   2   2   2   2   2   2   2   2   2   1
#>  [19]   1   1   1   1   1   1   1   1   6   8   8  74  28  17   7   1   1   1
#>  [37]   1  49   1   1   1   1   1   1  87  48   1   1   1   1   1   1   1   1
#>  [55]   1 192   1   1   1   1   1   1   1  91   1   1   1   1   1  94   1   1
#>  [73]   1   1   1 127   1   1   1   1   1   1   1   1   1   1   1   1   1   1
#>  [91]   1   1   1   1   1   1   1   1   1   1
#> 
#> $lambda
#>   [1] 5.079267981 4.736942319 4.417688261 4.119950858 3.842279959 3.583323149
#>   [7] 3.341819162 3.116591735 2.906543883 2.710652553 2.527963643 2.357587355
#>  [13] 2.198693857 2.050509250 1.912311790 1.783428375 1.663231272 1.551135051
#>  [19] 1.446593740 1.349098163 1.258173462 1.173376782 1.094295114 1.020543286
#>  [25] 0.951762084 0.887616505 0.827794123 0.772003570 0.719973113 0.671449335
#>  [31] 0.626195897 0.583992389 0.544633257 0.507926798 0.473694232 0.441768826
#>  [37] 0.411995086 0.384227996 0.358332315 0.334181916 0.311659174 0.290654388
#>  [43] 0.271065255 0.252796364 0.235758735 0.219869386 0.205050925 0.191231179
#>  [49] 0.178342838 0.166323127 0.155113505 0.144659374 0.134909816 0.125817346
#>  [55] 0.117337678 0.109429511 0.102054329 0.095176208 0.088761650 0.082779412
#>  [61] 0.077200357 0.071997311 0.067144933 0.062619590 0.058399239 0.054463326
#>  [67] 0.050792680 0.047369423 0.044176883 0.041199509 0.038422800 0.035833231
#>  [73] 0.033418192 0.031165917 0.029065439 0.027106526 0.025279636 0.023575874
#>  [79] 0.021986939 0.020505093 0.019123118 0.017834284 0.016632313 0.015511351
#>  [85] 0.014465937 0.013490982 0.012581735 0.011733768 0.010942951 0.010205433
#>  [91] 0.009517621 0.008876165 0.008277941 0.007720036 0.007199731 0.006714493
#>  [97] 0.006261959 0.005839924 0.005446333 0.005079268
#> 
#> $penalty
#> [1] "MCP"
#> 
#> $family
#> [1] "gaussian"
#> 
#> $gamma
#> [1] 3
#> 
#> $alpha
#> [1] 1
#> 
#> $convex.min
#> [1] 29
#> 
#> $loss
#>   [1] 692.00952 588.00449 500.98881 428.51758 368.48001 319.05479 278.67145
#>   [8] 245.97676 219.80560 199.15560 183.16526 171.09489 162.31009 156.26744
#>  [15] 152.50206 150.61691 150.23130 150.23130 150.23130 150.23130 150.23130
#>  [22] 150.23130 150.23130 150.23130 150.23130 150.23130 149.81415 144.55868
#>  [29] 140.32973 103.52332  92.88270  92.41149  92.40486  92.40485  92.40485
#>  [36]  92.40485  92.40485  82.25050  82.25049  82.25049  82.25048  82.25048
#>  [43]  82.25048  82.25048  78.56312  71.30660  71.30659  71.30658  71.30658
#>  [50]  71.30657  71.30657  71.30657  71.30657  71.30657  71.30657  60.38813
#>  [57]  60.38811  60.38810  60.38809  60.38808  60.38808  60.38807  60.38806
#>  [64]  59.36274  59.36273  59.36271  59.36269  59.36268  59.36266  59.06377
#>  [71]  59.06375  59.06373  59.06372  59.06370  59.06369  58.65456  58.65454
#>  [78]  58.65452  58.65451  58.65449  58.65448  58.65446  58.65445  58.65444
#>  [85]  58.65443  58.65442  58.65441  58.65440  58.65439  58.65438  58.65437
#>  [92]  58.65436  58.65436  58.65435  58.65434  58.65434  58.65433  58.65433
#>  [99]  58.65432  58.65432
#> 
#> $linear.predictors
#>     5.0793   4.7369   4.4177   4.1200   3.8423   3.5833   3.3418   3.1166
#> 1  18.2619 18.69514 19.09918 19.47598 19.82739 20.15512 20.46075 20.74579
#> 2  18.2619 18.91043 19.51524 20.07929 20.60532 21.09591 21.55343 21.98011
#> 3  18.2619 17.86711 17.49893 17.15556 16.83534 16.53669 16.25817 15.99843
#> 4  18.2619 18.42603 18.57910 18.72184 18.85497 18.97913 19.09492 19.20290
#> 5  18.2619 17.86711 17.49893 17.15556 16.83534 16.53669 16.25817 15.99843
#> 6  18.2619 18.66367 19.03837 19.38780 19.71369 20.01762 20.30106 20.56539
#> 7  18.2619 18.21571 18.17263 18.13246 18.09499 18.06005 18.02746 17.99707
#> 8  18.2619 18.21571 18.17263 18.13246 18.09499 18.06005 18.02746 17.99707
#> 9  18.2619 18.21571 18.17263 18.13246 18.09499 18.06005 18.02746 17.99707
#> 10 18.2619 17.40342 16.60279 15.85613 15.15979 14.51037 13.90473 13.33990
#> 11 18.2619 17.45310 16.69881 15.99535 15.33931 14.72748 14.15688 13.62475
#> 12 18.2619 17.53590 16.85883 16.22740 15.63851 15.08932 14.57714 14.09948
#> 13 18.2619 19.03173 19.74967 20.41923 21.04366 21.62601 22.16910 22.67560
#> 14 18.2619 19.04415 19.77368 20.45404 21.08854 21.68028 22.23214 22.74681
#> 15 18.2619 18.86033 19.41842 19.93890 20.42431 20.87699 21.29917 21.69289
#> 16 18.2619 18.04100 17.83498 17.64285 17.46367 17.29656 17.14072 16.99538
#> 17 18.2619 18.09896 17.94700 17.80528 17.67311 17.54985 17.43490 17.32769
#> 18 18.2619 17.90852 17.57894 17.27158 16.98494 16.71761 16.46830 16.23580
#> 19 18.2619 19.03049 19.74727 20.41575 21.03917 21.62058 22.16280 22.66848
#> 20 18.2619 17.90438 17.57094 17.25998 16.96998 16.69952 16.44729 16.21206
#> 21 18.2619 18.11138 17.97100 17.84009 17.71799 17.60413 17.49794 17.39890
#>      2.9065   2.7107   2.5280   2.3576   2.1987   2.0505   1.9123    1.7834
#> 1  21.01162 21.25954 21.49074 21.70636 21.90745 22.09499 22.26989 22.432995
#> 2  22.37804 22.74914 23.09524 23.41801 23.71903 23.99976 24.26156 24.505729
#> 3  15.75619 15.53028 15.31959 15.12310 14.93985 14.76896 14.60958 14.460945
#> 4  19.30361 19.39753 19.48512 19.56680 19.64298 19.71403 19.78029 19.842079
#> 5  15.75619 15.53028 15.31959 15.12310 14.93985 14.76896 14.60958 14.460945
#> 6  20.81192 21.04182 21.25624 21.45620 21.64268 21.81660 21.97879 22.130057
#> 7  17.96873 17.94229 17.91764 17.89465 17.87321 17.85322 17.83457 17.817178
#> 8  17.96873 17.94229 17.91764 17.89465 17.87321 17.85322 17.83457 17.817178
#> 9  17.96873 17.94229 17.91764 17.89465 17.87321 17.85322 17.83457 17.817178
#> 10 12.81315 12.32189 11.86374 11.43647 11.03800 10.66638 10.31981  9.996597
#> 11 13.12847 12.66565 12.23401 11.83147 11.45606 11.10594 10.77943 10.474920
#> 12 13.65402 13.23857 12.85113 12.48979 12.15282 11.83855 11.54546 11.272125
#> 13 23.14796 23.58848 23.99931 24.38246 24.73978 25.07302 25.38380 25.673634
#> 14 23.22679 23.67442 24.09188 24.48121 24.84429 25.18291 25.49870 25.793215
#> 15 22.06008 22.40252 22.72188 23.01972 23.29749 23.55653 23.79812 24.023419
#> 16 16.85983 16.73342 16.61553 16.50558 16.40305 16.30742 16.21825 16.135076
#> 17 17.22771 17.13447 17.04751 16.96641 16.89078 16.82025 16.75447 16.693119
#> 18 16.01896 15.81674 15.62815 15.45226 15.28823 15.13526 14.99260 14.859548
#> 19 23.14007 23.57989 23.99006 24.37258 24.72933 25.06203 25.37231 25.661676
#> 20 15.99268 15.78809 15.59729 15.41935 15.25340 15.09863 14.95429 14.819687
#> 21 17.30654 17.22041 17.14008 17.06516 16.99530 16.93014 16.86937 16.812700
#>       1.6632    1.5511    1.4466    1.3491    1.2582    1.1734    1.0943
#> 1  22.547325 22.547325 22.547325 22.547325 22.547325 22.547325 22.547325
#> 2  24.676872 24.676872 24.676872 24.676872 24.676872 24.676872 24.676872
#> 3  14.356761 14.356761 14.356761 14.356761 14.356761 14.356761 14.356761
#> 4  19.885392 19.885392 19.885392 19.885392 19.885392 19.885392 19.885392
#> 5  14.356761 14.356761 14.356761 14.356761 14.356761 14.356761 14.356761
#> 6  22.236084 22.236084 22.236084 22.236084 22.236084 22.236084 22.236084
#> 7  17.804988 17.804988 17.804988 17.804988 17.804988 17.804988 17.804988
#> 8  17.804988 17.804988 17.804988 17.804988 17.804988 17.804988 17.804988
#> 9  17.804988 17.804988 17.804988 17.804988 17.804988 17.804988 17.804988
#> 10  9.770044  9.770044  9.770044  9.770044  9.770044  9.770044  9.770044
#> 11 10.261478 10.261478 10.261478 10.261478 10.261478 10.261478 10.261478
#> 12 11.080535 11.080535 11.080535 11.080535 11.080535 11.080535 11.080535
#> 13 25.876790 25.876790 25.876790 25.876790 25.876790 25.876790 25.876790
#> 14 25.999648 25.999648 25.999648 25.999648 25.999648 25.999648 25.999648
#> 15 24.181343 24.181343 24.181343 24.181343 24.181343 24.181343 24.181343
#> 16 16.076779 16.076779 16.076779 16.076779 16.076779 16.076779 16.076779
#> 17 16.650119 16.650119 16.650119 16.650119 16.650119 16.650119 16.650119
#> 18 14.766289 14.766289 14.766289 14.766289 14.766289 14.766289 14.766289
#> 19 25.864504 25.864504 25.864504 25.864504 25.864504 25.864504 25.864504
#> 20 14.725336 14.725336 14.725336 14.725336 14.725336 14.725336 14.725336
#> 21 16.772977 16.772977 16.772977 16.772977 16.772977 16.772977 16.772977
#>       1.0205    0.9518    0.8876    0.8278    0.7720    0.7200   0.6714
#> 1  22.547325 22.547325 22.547325 22.558527 22.710491 22.852397 22.80469
#> 2  24.676872 24.676872 24.676872 24.685210 24.798325 24.903954 26.01116
#> 3  14.356761 14.356761 14.356761 14.353530 14.309699 14.268770 14.86050
#> 4  19.885392 19.885392 19.885392 19.874726 19.730039 19.594928 19.63221
#> 5  14.356761 14.356761 14.356761 14.353530 14.309699 14.268770 13.81379
#> 6  22.236084 22.236084 22.236084 22.222257 22.034687 21.859531 19.86848
#> 7  17.804988 17.804988 17.804988 17.797121 17.690385 17.590714 16.98686
#> 8  17.804988 17.804988 17.804988 17.797121 17.690385 17.590714 16.98686
#> 9  17.804988 17.804988 17.804988 17.797121 17.690385 17.590714 16.98686
#> 10  9.770044  9.770044  9.770044  9.772981  9.812825  9.850031 10.28923
#> 11 10.261478 10.261478 10.261478 10.263754 10.294633 10.323467 10.66687
#> 12 11.080535 11.080535 11.080535 11.081710 11.097646 11.112528 11.29625
#> 13 25.876790 25.876790 25.876790 25.883514 25.974740 26.059928 26.93321
#> 14 25.999648 25.999648 25.999648 26.006207 26.095192 26.178287 26.50426
#> 15 24.181343 24.181343 24.181343 24.164901 23.941844 23.733550 22.93334
#> 16 16.076779 16.076779 16.076779 16.071235 15.996027 15.925797 16.18221
#> 17 16.650119 16.650119 16.650119 16.643804 16.558136 16.478139 16.62278
#> 18 14.766289 14.766289 14.766289 14.762507 14.711206 14.663300 14.12848
#> 19 25.864504 25.864504 25.864504 25.871245 25.962695 26.048092 26.92377
#> 20 14.725336 14.725336 14.725336 14.747056 15.041714 15.316870 16.79406
#> 21 16.772977 16.772977 16.772977 16.791944 17.049247 17.289521 16.27410
#>      0.6262   0.5840   0.5446   0.5079   0.4737   0.4418   0.4120   0.3842
#> 1  22.71645 22.60445 22.58985 22.58983 22.58983 22.58982 22.58982 22.08477
#> 2  26.84046 26.99001 27.00952 27.00980 27.00997 27.01007 27.01013 26.45941
#> 3  15.33863 15.48874 15.50832 15.50847 15.50856 15.50861 15.50865 15.67262
#> 4  19.70986 19.81484 19.82852 19.82853 19.82854 19.82854 19.82854 18.71330
#> 5  13.47213 13.40889 13.40064 13.40053 13.40046 13.40041 13.40039 13.51425
#> 6  18.37187 18.09229 18.05581 18.05532 18.05501 18.05482 18.05471 18.92165
#> 7  16.54966 16.49841 16.49172 16.49157 16.49148 16.49142 16.49139 16.06142
#> 8  16.54966 16.49841 16.49172 16.49157 16.49148 16.49142 16.49139 16.06142
#> 9  16.54966 16.49841 16.49172 16.49157 16.49148 16.49142 16.49139 16.06142
#> 10 10.61988 10.68258 10.69077 10.69088 10.69094 10.69098 10.69101 10.65117
#> 11 10.92548 10.97469 10.98111 10.98119 10.98125 10.98128 10.98130 11.04335
#> 12 11.43481 11.46153 11.46502 11.46506 11.46509 11.46511 11.46512 11.93901
#> 13 27.58663 27.70323 27.71844 27.71866 27.71879 27.71888 27.71893 27.54316
#> 14 26.72978 26.73633 26.73719 26.73727 26.73732 26.73735 26.73737 28.27578
#> 15 22.38129 22.36831 22.36661 22.36641 22.36629 22.36621 22.36617 22.84224
#> 16 16.40822 16.51110 16.52452 16.52458 16.52462 16.52465 16.52466 15.68918
#> 17 16.76475 16.85189 16.86325 16.86329 16.86331 16.86333 16.86334 16.80518
#> 18 13.72679 13.65231 13.64259 13.64246 13.64238 13.64233 13.64230 14.82697
#> 19 27.57899 27.69592 27.71119 27.71140 27.71154 27.71162 27.71167 27.53709
#> 20 17.85235 17.95512 17.96855 17.96891 17.96914 17.96928 17.96936 18.90528
#> 21 15.39266 15.01252 14.96295 14.96270 14.96254 14.96245 14.96239 13.89133
#>      0.3583   0.3342   0.3117   0.2907   0.2711   0.2528   0.2358   0.2199
#> 1  22.08471 22.08471 22.08474 22.08477 22.08479 22.08482 21.83409 22.79988
#> 2  26.45975 26.46001 26.46020 26.46033 26.46042 26.46048 26.45898 26.40503
#> 3  15.67285 15.67298 15.67306 15.67309 15.67311 15.67311 15.33041 15.08847
#> 4  18.71338 18.71346 18.71353 18.71358 18.71362 18.71364 19.13624 18.57867
#> 5  13.51412 13.51402 13.51395 13.51390 13.51387 13.51384 13.00945 13.50477
#> 6  18.92093 18.92044 18.92012 18.91990 18.91977 18.91969 19.31112 17.98691
#> 7  16.06125 16.06115 16.06109 16.06105 16.06103 16.06102 15.78046 15.98326
#> 8  16.06125 16.06115 16.06109 16.06105 16.06103 16.06102 15.87322 16.07531
#> 9  16.06125 16.06115 16.06109 16.06105 16.06103 16.06102 16.05872 16.25942
#> 10 10.65137 10.65148 10.65154 10.65157 10.65159 10.65159 11.46896 11.69939
#> 11 11.04351 11.04360 11.04364 11.04366 11.04366 11.04366 11.75895 12.03174
#> 12 11.93909 11.93912 11.93912 11.93911 11.93909 11.93907 12.44335 12.82754
#> 13 27.54340 27.54359 27.54372 27.54382 27.54388 27.54392 27.88413 27.95025
#> 14 28.27576 28.27573 28.27569 28.27566 28.27564 28.27562 28.50955 29.24146
#> 15 22.84195 22.84175 22.84161 22.84152 22.84146 22.84143 22.98362 22.90894
#> 16 15.68934 15.68945 15.68953 15.68959 15.68962 15.68965 15.06230 14.74133
#> 17 16.80526 16.80531 16.80533 16.80534 16.80534 16.80534 16.40194 16.23792
#> 18 14.82677 14.82659 14.82645 14.82634 14.82627 14.82621 14.18459 14.88047
#> 19 27.53733 27.53752 27.53766 27.53775 27.53782 27.53786 27.61473 27.68243
#> 20 18.90571 18.90598 18.90614 18.90623 18.90627 18.90630 18.45572 16.94065
#> 21 13.89098 13.89080 13.89072 13.89069 13.89069 13.89070 13.93944 13.67618
#>      0.2051   0.1912   0.1783   0.1663   0.1551   0.1447   0.1349   0.1258
#> 1  22.80023 22.80053 22.80078 22.80100 22.80118 22.80132 22.80145 22.80155
#> 2  26.40510 26.40514 26.40517 26.40518 26.40519 26.40519 26.40519 26.40519
#> 3  15.08831 15.08817 15.08804 15.08793 15.08784 15.08776 15.08769 15.08763
#> 4  18.57849 18.57833 18.57818 18.57805 18.57794 18.57784 18.57776 18.57769
#> 5  13.50482 13.50486 13.50490 13.50493 13.50496 13.50498 13.50500 13.50501
#> 6  17.98651 17.98619 17.98594 17.98574 17.98557 17.98544 17.98533 17.98525
#> 7  15.98323 15.98321 15.98319 15.98317 15.98315 15.98314 15.98312 15.98311
#> 8  16.07529 16.07528 16.07527 16.07526 16.07525 16.07524 16.07523 16.07522
#> 9  16.25942 16.25942 16.25943 16.25943 16.25943 16.25944 16.25944 16.25944
#> 10 11.69953 11.69964 11.69974 11.69983 11.69991 11.69998 11.70004 11.70009
#> 11 12.03188 12.03199 12.03209 12.03218 12.03226 12.03233 12.03239 12.03244
#> 12 12.82768 12.82780 12.82791 12.82800 12.82809 12.82816 12.82823 12.82829
#> 13 27.95038 27.95049 27.95057 27.95064 27.95070 27.95074 27.95078 27.95081
#> 14 29.24176 29.24204 29.24228 29.24249 29.24267 29.24283 29.24296 29.24308
#> 15 22.90884 22.90876 22.90870 22.90866 22.90862 22.90860 22.90857 22.90855
#> 16 14.74112 14.74093 14.74075 14.74059 14.74046 14.74033 14.74023 14.74014
#> 17 16.23777 16.23764 16.23752 16.23742 16.23733 16.23725 16.23719 16.23713
#> 18 14.88055 14.88063 14.88070 14.88077 14.88084 14.88089 14.88094 14.88098
#> 19 27.68253 27.68261 27.68268 27.68272 27.68276 27.68279 27.68282 27.68283
#> 20 16.94031 16.94000 16.93974 16.93951 16.93932 16.93916 16.93902 16.93891
#> 21 13.67627 13.67635 13.67642 13.67649 13.67654 13.67659 13.67663 13.67666
#>      0.1173   0.1094   0.1021   0.0952   0.0888   0.0828   0.0772   0.0720
#> 1  22.80163 22.62815 22.62823 22.62831 22.62840 22.62848 22.62856 22.62864
#> 2  26.40519 24.90527 24.90546 24.90563 24.90579 24.90594 24.90608 24.90621
#> 3  15.08758 16.04741 16.04724 16.04708 16.04692 16.04676 16.04661 16.04645
#> 4  18.57763 18.97331 18.97326 18.97321 18.97316 18.97310 18.97305 18.97300
#> 5  13.50503 13.48845 13.48860 13.48873 13.48885 13.48896 13.48907 13.48916
#> 6  17.98517 17.66170 17.66179 17.66189 17.66200 17.66211 17.66222 17.66234
#> 7  15.98309 15.50307 15.50309 15.50312 15.50315 15.50319 15.50323 15.50328
#> 8  16.07521 16.76600 16.76589 16.76579 16.76570 16.76562 16.76555 16.76549
#> 9  16.25944 17.47495 17.47475 17.47457 17.47440 17.47424 17.47410 17.47397
#> 10 11.70014 12.02026 12.02019 12.02012 12.02005 12.01998 12.01990 12.01982
#> 11 12.03249 11.58959 11.58960 11.58961 11.58961 11.58961 11.58961 11.58961
#> 12 12.82834 11.93386 11.93392 11.93397 11.93402 11.93407 11.93412 11.93417
#> 13 27.95084 28.03525 28.03517 28.03510 28.03503 28.03496 28.03490 28.03484
#> 14 29.24318 30.69164 30.69144 30.69125 30.69107 30.69090 30.69074 30.69059
#> 15 22.90853 22.07727 22.07755 22.07781 22.07806 22.07829 22.07851 22.07872
#> 16 14.74006 14.34961 14.34966 14.34971 14.34975 14.34978 14.34981 14.34984
#> 17 16.23708 16.99535 16.99519 16.99504 16.99489 16.99475 16.99462 16.99449
#> 18 14.88101 13.91711 13.91729 13.91747 13.91765 13.91782 13.91798 13.91813
#> 19 27.68285 27.50535 27.50535 27.50534 27.50534 27.50534 27.50533 27.50532
#> 20 16.93882 16.96893 16.96874 16.96855 16.96836 16.96818 16.96800 16.96783
#> 21 13.67668 13.96746 13.96758 13.96769 13.96780 13.96790 13.96800 13.96810
#>      0.0671   0.0626   0.0584   0.0545   0.0508   0.0474   0.0442   0.0412
#> 1  22.62872 22.17797 22.17770 22.17743 22.17717 22.17692 22.17668 21.97854
#> 2  24.90633 24.85039 24.85027 24.85016 24.85005 24.84995 24.84986 25.03440
#> 3  16.04630 15.98610 15.98621 15.98632 15.98642 15.98652 15.98662 15.94934
#> 4  18.97294 19.05889 19.05896 19.05904 19.05911 19.05918 19.05925 19.05208
#> 5  13.48924 13.67132 13.67134 13.67137 13.67140 13.67143 13.67146 13.56964
#> 6  17.66246 17.32895 17.32862 17.32830 17.32798 17.32768 17.32738 17.25415
#> 7  15.50334 15.51440 15.51435 15.51430 15.51424 15.51419 15.51414 15.66757
#> 8  16.76543 16.94480 16.94493 16.94506 16.94517 16.94528 16.94539 16.96601
#> 9  17.47385 17.78545 17.78570 17.78594 17.78617 17.78639 17.78659 17.78515
#> 10 12.01975 11.92508 11.92512 11.92516 11.92520 11.92523 11.92527 11.76771
#> 11 11.58961 11.51914 11.51913 11.51911 11.51910 11.51908 11.51907 11.49667
#> 12 11.93421 11.97567 11.97566 11.97566 11.97565 11.97565 11.97565 12.07304
#> 13 28.03478 28.15663 28.15675 28.15686 28.15698 28.15708 28.15719 28.24204
#> 14 30.69044 30.74451 30.74468 30.74484 30.74499 30.74514 30.74527 30.56330
#> 15 22.07891 22.32523 22.32516 22.32511 22.32506 22.32501 22.32497 22.32424
#> 16 14.34986 13.89127 13.89102 13.89079 13.89056 13.89034 13.89012 13.90586
#> 17 16.99437 16.84584 16.84587 16.84590 16.84592 16.84594 16.84596 16.85821
#> 18 13.91828 14.03111 14.03104 14.03097 14.03091 14.03085 14.03080 14.12740
#> 19 27.50531 27.49269 27.49270 27.49271 27.49273 27.49274 27.49276 27.54364
#> 20 16.96767 16.89951 16.89962 16.89972 16.89982 16.89991 16.90000 17.14952
#> 21 13.96819 14.37506 14.37516 14.37527 14.37537 14.37547 14.37557 14.19150
#>      0.0384   0.0358   0.0334   0.0312   0.0291   0.0271   0.0253   0.0236
#> 1  21.97879 21.97903 21.97925 21.97946 21.97967 21.80707 21.80691 21.80675
#> 2  25.03444 25.03449 25.03453 25.03456 25.03459 25.11517 25.11516 25.11516
#> 3  15.94920 15.94907 15.94896 15.94885 15.94874 15.88147 15.88148 15.88148
#> 4  19.05198 19.05189 19.05180 19.05172 19.05165 19.34972 19.34990 19.35008
#> 5  13.56939 13.56917 13.56895 13.56875 13.56857 13.68627 13.68638 13.68648
#> 6  17.25462 17.25508 17.25551 17.25592 17.25631 17.35232 17.35213 17.35194
#> 7  15.66779 15.66800 15.66819 15.66837 15.66853 15.66904 15.66903 15.66903
#> 8  16.96600 16.96599 16.96598 16.96596 16.96595 16.89803 16.89806 16.89810
#> 9  17.78505 17.78495 17.78485 17.78476 17.78467 17.64815 17.64820 17.64824
#> 10 11.76757 11.76744 11.76731 11.76720 11.76710 11.71131 11.71130 11.71130
#> 11 11.49664 11.49661 11.49659 11.49657 11.49655 11.46545 11.46543 11.46541
#> 12 12.07305 12.07307 12.07308 12.07309 12.07310 12.07217 12.07217 12.07218
#> 13 28.24198 28.24192 28.24186 28.24181 28.24175 28.21428 28.21432 28.21437
#> 14 30.56317 30.56305 30.56293 30.56282 30.56272 30.70940 30.70960 30.70980
#> 15 22.32410 22.32396 22.32384 22.32372 22.32361 21.96419 21.96390 21.96360
#> 16 13.90609 13.90631 13.90652 13.90672 13.90692 13.96045 13.96036 13.96028
#> 17 16.85828 16.85835 16.85842 16.85848 16.85854 16.81436 16.81434 16.81433
#> 18 14.12741 14.12742 14.12743 14.12743 14.12744 14.33150 14.33162 14.33174
#> 19 27.54362 27.54359 27.54357 27.54355 27.54353 27.59559 27.59563 27.59568
#> 20 17.14957 17.14961 17.14965 17.14969 17.14972 17.03906 17.03898 17.03891
#> 21 14.19124 14.19100 14.19077 14.19056 14.19035 14.21501 14.21508 14.21515
#>      0.0220   0.0205   0.0191   0.0178   0.0166   0.0155   0.0145   0.0135
#> 1  21.80660 21.80646 21.80631 21.80617 21.80604 21.80591 21.80578 21.80566
#> 2  25.11515 25.11514 25.11514 25.11513 25.11513 25.11512 25.11512 25.11511
#> 3  15.88148 15.88148 15.88148 15.88148 15.88148 15.88148 15.88148 15.88148
#> 4  19.35025 19.35042 19.35058 19.35074 19.35089 19.35104 19.35118 19.35132
#> 5  13.68658 13.68668 13.68677 13.68686 13.68695 13.68704 13.68712 13.68720
#> 6  17.35176 17.35158 17.35141 17.35125 17.35109 17.35093 17.35078 17.35063
#> 7  15.66902 15.66901 15.66900 15.66900 15.66899 15.66899 15.66898 15.66898
#> 8  16.89813 16.89816 16.89819 16.89822 16.89825 16.89828 16.89831 16.89834
#> 9  17.64829 17.64833 17.64838 17.64842 17.64846 17.64850 17.64854 17.64858
#> 10 11.71129 11.71128 11.71127 11.71126 11.71125 11.71124 11.71124 11.71123
#> 11 11.46540 11.46538 11.46536 11.46535 11.46533 11.46531 11.46530 11.46528
#> 12 12.07219 12.07219 12.07220 12.07221 12.07221 12.07222 12.07223 12.07223
#> 13 28.21441 28.21445 28.21449 28.21453 28.21457 28.21461 28.21465 28.21469
#> 14 30.70999 30.71018 30.71036 30.71053 30.71070 30.71087 30.71103 30.71118
#> 15 21.96332 21.96305 21.96278 21.96253 21.96228 21.96204 21.96180 21.96158
#> 16 13.96020 13.96012 13.96004 13.95996 13.95989 13.95982 13.95975 13.95968
#> 17 16.81431 16.81430 16.81429 16.81427 16.81426 16.81425 16.81424 16.81423
#> 18 14.33185 14.33196 14.33207 14.33217 14.33227 14.33237 14.33246 14.33256
#> 19 27.59572 27.59576 27.59580 27.59584 27.59588 27.59592 27.59596 27.59599
#> 20 17.03884 17.03877 17.03871 17.03864 17.03858 17.03852 17.03847 17.03841
#> 21 14.21522 14.21529 14.21536 14.21542 14.21548 14.21555 14.21560 14.21566
#>      0.0126   0.0117   0.0109   0.0102   0.0095   0.0089   0.0083   0.0077
#> 1  21.80554 21.80542 21.80530 21.80519 21.80509 21.80498 21.80488 21.80478
#> 2  25.11511 25.11510 25.11510 25.11509 25.11509 25.11508 25.11508 25.11508
#> 3  15.88148 15.88148 15.88148 15.88148 15.88148 15.88148 15.88148 15.88148
#> 4  19.35145 19.35158 19.35171 19.35183 19.35195 19.35206 19.35218 19.35228
#> 5  13.68728 13.68735 13.68742 13.68749 13.68756 13.68763 13.68769 13.68775
#> 6  17.35048 17.35034 17.35021 17.35007 17.34995 17.34982 17.34970 17.34958
#> 7  15.66897 15.66897 15.66896 15.66896 15.66895 15.66895 15.66895 15.66895
#> 8  16.89836 16.89839 16.89842 16.89844 16.89847 16.89849 16.89851 16.89854
#> 9  17.64862 17.64865 17.64869 17.64873 17.64876 17.64879 17.64883 17.64886
#> 10 11.71122 11.71121 11.71120 11.71120 11.71119 11.71118 11.71118 11.71117
#> 11 11.46527 11.46526 11.46524 11.46523 11.46522 11.46520 11.46519 11.46518
#> 12 12.07224 12.07224 12.07225 12.07226 12.07226 12.07227 12.07227 12.07228
#> 13 28.21472 28.21476 28.21479 28.21482 28.21486 28.21489 28.21492 28.21495
#> 14 30.71133 30.71148 30.71162 30.71176 30.71189 30.71202 30.71215 30.71227
#> 15 21.96136 21.96114 21.96094 21.96074 21.96054 21.96035 21.96017 21.95999
#> 16 13.95961 13.95954 13.95948 13.95942 13.95936 13.95930 13.95924 13.95918
#> 17 16.81422 16.81421 16.81420 16.81419 16.81418 16.81417 16.81416 16.81415
#> 18 14.33265 14.33273 14.33282 14.33290 14.33298 14.33305 14.33313 14.33320
#> 19 27.59602 27.59606 27.59609 27.59612 27.59615 27.59618 27.59621 27.59624
#> 20 17.03836 17.03831 17.03826 17.03821 17.03816 17.03812 17.03808 17.03803
#> 21 14.21572 14.21577 14.21582 14.21587 14.21592 14.21597 14.21602 14.21606
#>      0.0072   0.0067   0.0063   0.0058   0.0054   0.0051
#> 1  21.80468 21.80459 21.80450 21.80441 21.80432 21.80424
#> 2  25.11507 25.11507 25.11506 25.11506 25.11506 25.11505
#> 3  15.88148 15.88148 15.88148 15.88148 15.88148 15.88148
#> 4  19.35239 19.35249 19.35259 19.35268 19.35278 19.35286
#> 5  13.68781 13.68787 13.68793 13.68798 13.68803 13.68808
#> 6  17.34947 17.34936 17.34925 17.34914 17.34904 17.34894
#> 7  15.66894 15.66894 15.66894 15.66894 15.66893 15.66893
#> 8  16.89856 16.89858 16.89860 16.89863 16.89865 16.89867
#> 9  17.64889 17.64892 17.64895 17.64898 17.64901 17.64904
#> 10 11.71116 11.71115 11.71115 11.71114 11.71114 11.71113
#> 11 11.46517 11.46516 11.46515 11.46514 11.46513 11.46512
#> 12 12.07228 12.07229 12.07229 12.07230 12.07230 12.07231
#> 13 28.21498 28.21500 28.21503 28.21506 28.21508 28.21511
#> 14 30.71239 30.71250 30.71262 30.71272 30.71283 30.71293
#> 15 21.95982 21.95965 21.95949 21.95934 21.95918 21.95904
#> 16 13.95913 13.95907 13.95902 13.95897 13.95892 13.95887
#> 17 16.81414 16.81413 16.81412 16.81412 16.81411 16.81410
#> 18 14.33327 14.33334 14.33340 14.33347 14.33353 14.33359
#> 19 27.59626 27.59629 27.59631 27.59634 27.59636 27.59638
#> 20 17.03799 17.03795 17.03792 17.03788 17.03784 17.03781
#> 21 14.21611 14.21615 14.21619 14.21623 14.21627 14.21631
#> 
#> $penalty.factor
#>  [1] 1 1 1 1 1 1 1 1 1 1
#> 
#> $n
#> [1] 21
#> 
#> $y
#>  [1] 21.0 22.8 18.7 18.1 14.3 17.8 16.4 17.3 15.2 10.4 10.4 14.7 32.4 30.4 21.5
#> [16] 15.5 15.2 13.3 27.3 15.8 15.0
#> 
#> $X
#>               am        carb        cyl        disp       drat       gear
#>  [1,]  1.4142136  0.63614643 -0.4514661 -0.84370826  0.7366317  0.7897540
#>  [2,]  1.4142136 -1.18554561 -1.6365647 -1.26297080  0.6461682  0.7897540
#>  [3,] -0.7071068 -0.57831493  0.7336324  0.76883995 -0.6203214 -0.7179582
#>  [4,] -0.7071068 -1.18554561 -0.4514661 -0.31963009 -1.3259370 -0.7179582
#>  [5,] -0.7071068  0.63614643  0.7336324  0.76883995 -0.5117652 -0.7179582
#>  [6,] -0.7071068  0.63614643 -0.4514661 -0.78243143  0.7728171  0.7897540
#>  [7,] -0.7071068  0.02891575  0.7336324  0.08995715 -0.7650631 -0.7179582
#>  [8,] -0.7071068  0.02891575  0.7336324  0.08995715 -0.7650631 -0.7179582
#>  [9,] -0.7071068  0.02891575  0.7336324  0.08995715 -0.7650631 -0.7179582
#> [10,] -0.7071068  0.63614643  0.7336324  1.67186695 -1.0183610 -0.7179582
#> [11,] -0.7071068  0.63614643  0.7336324  1.57511406 -0.8917120 -0.7179582
#> [12,] -0.7071068  0.63614643  0.7336324  1.41385924 -0.4755798 -0.7179582
#> [13,]  1.4142136 -1.18554561 -1.6365647 -1.49920911  1.0623004  0.7897540
#> [14,]  1.4142136 -0.57831493 -1.6365647 -1.52339733  2.6001806  0.7897540
#> [15,] -0.7071068 -1.18554561 -1.6365647 -1.16541163  0.3747775 -0.7179582
#> [16,] -0.7071068 -0.57831493  0.7336324  0.43020483 -1.3259370 -0.7179582
#> [17,] -0.7071068 -0.57831493  0.7336324  0.31732645 -0.6203214 -0.7179582
#> [18,] -0.7071068  0.63614643  0.7336324  0.68821254  0.4290557 -0.7179582
#> [19,]  1.4142136 -1.18554561 -1.6365647 -1.49679029  1.0623004  0.7897540
#> [20,]  1.4142136  0.63614643  0.7336324  0.69627528  1.3155984  2.2974661
#> [21,]  1.4142136  3.06506914  0.7336324  0.29313823  0.0852942  2.2974661
#>               hp        qsec         vs           wt
#>  [1,] -0.7550561 -0.71156559 -0.7071068 -0.804335651
#>  [2,] -0.9882352  0.67382407  1.4142136 -1.099307104
#>  [3,]  0.1365110 -0.35071991 -0.7071068  0.001919655
#>  [4,] -0.8236382  1.71125540  1.4142136  0.021584419
#>  [5,]  1.0966602 -1.11107330 -0.7071068  0.129740619
#>  [6,] -0.5767427  0.86069059  1.4142136  0.001919655
#>  [7,]  0.2050931 -0.10586034 -0.7071068  0.621359708
#>  [8,]  0.2050931  0.02301312 -0.7071068  0.287058727
#>  [9,]  0.2050931  0.28076003 -0.7071068  0.336220636
#> [10,]  0.5480035  0.26787269 -0.7071068  1.781580758
#> [11,]  0.6851677  0.16477392 -0.7071068  1.952664201
#> [12,]  0.8909139 -0.09297299 -0.7071068  1.874988385
#> [13,] -1.3585784  1.22797994  1.4142136 -1.217295686
#> [14,] -1.5506083  0.61583102  1.4142136 -1.792490020
#> [15,] -0.9333695  1.57593827  1.4142136 -0.956737568
#> [16,] -0.2063994 -0.44737500 -0.7071068  0.080578710
#> [17,] -0.2063994 -0.17029707 -0.7071068 -0.002996535
#> [18,]  1.0966602 -1.38815124 -0.7071068  0.395214927
#> [19,] -1.3585784  0.86069059  1.4142136 -1.477853803
#> [20,]  1.3572721 -1.97452547 -0.7071068 -0.263554653
#> [21,]  2.3311377 -1.91008874 -0.7071068  0.129740619
#> attr(,"center")
#>  [1]   0.3333333   2.9523810   6.7619048 264.6428571   3.4928571   3.4761905
#>  [7] 165.0476190  17.5642857   0.3333333   3.4380476
#> attr(,"scale")
#>  [1]   0.4714045   1.6468206   1.6876233 124.0272995   0.5527089   0.6632566
#>  [7]  72.9053373   1.5519099   0.4714045   1.0170476
#> attr(,"nonsingular")
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> attr(,"class")
#> [1] "ncvreg"
#> 
#> $min
#> [1] 80
#> 
#> $lambda.min
#> [1] 0.02050509
#> 
#> $null.dev
#> [1] 32.95283
#> 
#> $Bias
#> [1] 7.718672
#> 
#> attr(,"class")
#> [1] "cv.ncvreg"

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 33.88493