Regression MultilayerPerceptron Learner

Regressor that uses backpropagation to learn a multi-layer perceptron. Calls RWeka::make_Weka_classifier() from RWeka.

Custom mlr3 parameters

output_debug_info:

original id: output-debug-info

do_not_check_capabilities:

original id: do-not-check-capabilities

num_decimal_places:

original id: num-decimal-places

batch_size:

original id: batch-size

Reason for change: This learner contains changed ids of the following control arguments since their ids contain irregular pattern

G removed:

GUI will be opened

Reason for change: The parameter is removed because we don't want to launch GUI.

Dictionary

This Learner can be instantiated via lrn():

lrn("regr.multilayer_perceptron")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, RWeka

Parameters

Id	Type	Default	Levels	Range

subset	untyped	-		-

na.action	untyped	-		-

L	numeric	0.3		\([0, 1]\)

M	numeric	0.2		\([0, 1]\)

N	integer	500		\([1, \infty)\)

V	numeric	0		\([0, 100]\)

S	integer	0		\([0, \infty)\)

E	integer	20		\([1, \infty)\)

A	logical	FALSE	TRUE, FALSE	-

B	logical	FALSE	TRUE, FALSE	-

H	untyped	"a"		-

C	logical	FALSE	TRUE, FALSE	-

I	logical	FALSE	TRUE, FALSE	-

R	logical	FALSE	TRUE, FALSE	-

D	logical	FALSE	TRUE, FALSE	-

output_debug_info	logical	FALSE	TRUE, FALSE	-

do_not_check_capabilities	logical	FALSE	TRUE, FALSE	-

num_decimal_places	integer	2		\([1, \infty)\)

batch_size	integer	100		\([1, \infty)\)

options	untyped	NULL		-

See also

Author

damirpolat

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrMultilayerPerceptron

Methods

Public methods

LearnerRegrMultilayerPerceptron$new()

LearnerRegrMultilayerPerceptron$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrMultilayerPerceptron$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrMultilayerPerceptron$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("regr.multilayer_perceptron")
print(learner)
#> <LearnerRegrMultilayerPerceptron:regr.multilayer_perceptron>: MultilayerPerceptron
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, RWeka
#> * Predict Types:  [response]
#> * Feature Types: logical, integer, numeric, factor, ordered
#> * Properties: missings

# Define a Task
task = mlr3::tsk("mtcars")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> Linear Node 0
#>     Inputs    Weights
#>     Threshold    1.9678988248294351
#>     Node 1    -0.5265296587251189
#>     Node 2    -2.5269023614954884
#>     Node 3    0.7811952705523281
#>     Node 4    -0.5796549032104666
#>     Node 5    -0.5274527423489601
#> Sigmoid Node 1
#>     Inputs    Weights
#>     Threshold    -0.7708695481359399
#>     Attrib am    -0.33879200379088426
#>     Attrib carb    0.6566650275597203
#>     Attrib cyl    -0.44677615037604956
#>     Attrib disp    0.3165947271994374
#>     Attrib drat    0.26361835974735776
#>     Attrib gear    -0.048711865728405544
#>     Attrib hp    0.45816060021120997
#>     Attrib qsec    0.09465790553319157
#>     Attrib vs    0.5878625418459925
#>     Attrib wt    -0.22127964280453924
#> Sigmoid Node 2
#>     Inputs    Weights
#>     Threshold    2.967322110594636
#>     Attrib am    -0.1873910360853556
#>     Attrib carb    0.8841688220904842
#>     Attrib cyl    -0.39627673658056156
#>     Attrib disp    0.14974801000364024
#>     Attrib drat    0.18795884045086125
#>     Attrib gear    -0.2447872305365602
#>     Attrib hp    0.9678463363390377
#>     Attrib qsec    -0.9606855793569181
#>     Attrib vs    -0.16758520810011132
#>     Attrib wt    1.8075048196357322
#> Sigmoid Node 3
#>     Inputs    Weights
#>     Threshold    -1.0853551583622278
#>     Attrib am    0.36096583501810064
#>     Attrib carb    0.5296030414580487
#>     Attrib cyl    -0.3711361483932099
#>     Attrib disp    -0.5654793209201966
#>     Attrib drat    -0.16998470107410343
#>     Attrib gear    0.8583576027143862
#>     Attrib hp    -0.11691949781128083
#>     Attrib qsec    0.35577797991793203
#>     Attrib vs    0.3769030658960427
#>     Attrib wt    -0.19904410093209418
#> Sigmoid Node 4
#>     Inputs    Weights
#>     Threshold    -0.7696698064187772
#>     Attrib am    -0.28572961255125073
#>     Attrib carb    0.6633314523752937
#>     Attrib cyl    -0.451308183126236
#>     Attrib disp    0.34371469612252453
#>     Attrib drat    0.3425241561788499
#>     Attrib gear    -0.13940681693996199
#>     Attrib hp    0.5394755430599535
#>     Attrib qsec    0.02828744181313037
#>     Attrib vs    0.5388783341404154
#>     Attrib wt    -0.3165152837344715
#> Sigmoid Node 5
#>     Inputs    Weights
#>     Threshold    -0.5445267252132117
#>     Attrib am    -0.12221276062766391
#>     Attrib carb    0.57464168725051
#>     Attrib cyl    -0.35805438788463
#>     Attrib disp    0.2944120224648059
#>     Attrib drat    0.606346655437687
#>     Attrib gear    -0.005432107622801065
#>     Attrib hp    0.5476151582546486
#>     Attrib qsec    -0.13374957033513676
#>     Attrib vs    0.43419932926297816
#>     Attrib wt    -0.3438059448941365
#> Class 
#>     Input
#>     Node 0
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 13.47642