Accelerated Oblique Random Survival Forest Learner

Accelerated oblique random survival forest. Calls aorsf::orsf() from aorsf. Note that although the learner has the property "missing" and it can in principle deal with missing values, the behaviour has to be configured using the parameter na_action.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("surv.aorsf")
lrn("surv.aorsf")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, mlr3proba, mlr3extralearners, aorsf, pracma

Parameters

Id	Type	Default	Levels	Range

n_tree	integer	500		\([1, \infty)\)

n_split	integer	5		\([1, \infty)\)

n_retry	integer	3		\([0, \infty)\)

n_thread	integer	0		\([0, \infty)\)

pred_aggregate	logical	TRUE	TRUE, FALSE	-

pred_simplify	logical	FALSE	TRUE, FALSE	-

mtry	integer	NULL		\([1, \infty)\)

mtry_ratio	numeric	-		\([0, 1]\)

sample_with_replacement	logical	TRUE	TRUE, FALSE	-

sample_fraction	numeric	0.632		\([0, 1]\)

control_type	character	fast	fast, cph, net	-

split_rule	character	logrank	logrank, cstat	-

control_fast_do_scale	logical	FALSE	TRUE, FALSE	-

control_fast_ties	character	efron	efron, breslow	-

control_cph_ties	character	efron	efron, breslow	-

control_cph_eps	numeric	1e-09		\([0, \infty)\)

control_cph_iter_max	integer	20		\([1, \infty)\)

control_net_alpha	numeric	0.5		\((-\infty, \infty)\)

control_net_df_target	integer	NULL		\([1, \infty)\)

leaf_min_events	integer	1		\([1, \infty)\)

leaf_min_obs	integer	5		\([1, \infty)\)

split_min_events	integer	5		\([1, \infty)\)

split_min_obs	integer	10		\([1, \infty)\)

split_min_stat	numeric	NULL		\([0, \infty)\)

oobag_pred_type	character	surv	none, surv, risk, chf	-

importance	character	anova	none, anova, negate, permute	-

importance_max_pvalue	numeric	0.01		\([1e-04, 0.9999]\)

oobag_pred_horizon	numeric	NULL		\([0, \infty)\)

oobag_eval_every	integer	NULL		\([1, \infty)\)

attach_data	logical	TRUE	TRUE, FALSE	-

verbose_progress	logical	FALSE	TRUE, FALSE	-

na_action	character	fail	fail, omit, impute_meanmode	-

Initial parameter values

mtry:

This hyperparameter can alternatively be set via the added hyperparameter mtry_ratio as mtry = max(ceiling(mtry_ratio * n_features), 1). Note that mtry and mtry_ratio are mutually exclusive.

References

Jaeger BC, Long DL, Long DM, Sims M, Szychowski JM, Min Y, Mcclure LA, Howard G, Simon N (2019). “Oblique random survival forests.” The Annals of Applied Statistics, 13(3). doi:10.1214/19-aoas1261 .

Jaeger BC, Welden S, Lenoir K, Speiser JL, Segar MW, Pandey A, Pajewski NM (2023). “Accelerated and interpretable oblique random survival forests.” Journal of Computational and Graphical Statistics, 1--16. doi:10.1080/10618600.2023.2231048 .

See also

Author

bcjaeger

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvAorsf

Methods

Public methods

LearnerSurvAorsf$new()

LearnerSurvAorsf$oob_error()

LearnerSurvAorsf$importance()

LearnerSurvAorsf$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvAorsf$new()

Method oob_error()

OOB concordance error extracted from the model slot eval_oobag$stat_values

Usage

LearnerSurvAorsf$oob_error()

Returns

numeric().

Method importance()

The importance scores are extracted from the model.

Usage

LearnerSurvAorsf$importance()

Returns

Named numeric().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvAorsf$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

learner = mlr3::lrn("surv.aorsf")
print(learner)
#> <LearnerSurvAorsf:surv.aorsf>: Oblique Random Forest
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3proba, mlr3extralearners, aorsf, pracma
#> * Predict Types:  [crank], distr
#> * Feature Types: integer, numeric, factor, ordered
#> * Properties: importance, missings, oob_error

# available parameters:
learner$param_set$ids()
#>  [1] "n_tree"                  "n_split"                
#>  [3] "n_retry"                 "n_thread"               
#>  [5] "pred_aggregate"          "pred_simplify"          
#>  [7] "mtry"                    "mtry_ratio"             
#>  [9] "sample_with_replacement" "sample_fraction"        
#> [11] "control_type"            "split_rule"             
#> [13] "control_fast_do_scale"   "control_fast_ties"      
#> [15] "control_cph_ties"        "control_cph_eps"        
#> [17] "control_cph_iter_max"    "control_net_alpha"      
#> [19] "control_net_df_target"   "leaf_min_events"        
#> [21] "leaf_min_obs"            "split_min_events"       
#> [23] "split_min_obs"           "split_min_stat"         
#> [25] "oobag_pred_type"         "importance"             
#> [27] "importance_max_pvalue"   "oobag_pred_horizon"     
#> [29] "oobag_eval_every"        "attach_data"            
#> [31] "verbose_progress"        "na_action"