Cross-Validated GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::cv.glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.cv.glmnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.cv.glmnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.cv_glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

foldid	untyped	NULL		-

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nfolds	integer	10		\([3, \infty)\)

nlambda	integer	100		\([1, \infty)\)

offset	untyped	NULL		-

newoffset	untyped	-		-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	lambda.1se		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

standardize.response	logical	FALSE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.gaussian	character	-	covariance, naive	-

type.logistic	character	Newton	Newton, modified.Newton	-

type.measure	character	deviance	deviance, C	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvCVGlmnet

Methods

Public methods

LearnerSurvCVGlmnet$new()

LearnerSurvCVGlmnet$selected_features()

LearnerSurvCVGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvCVGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvCVGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvCVGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.cv_glmnet")
print(learner)
#> <LearnerSurvCVGlmnet:surv.cv_glmnet>: Regularized Generalized Linear Model
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3proba, mlr3extralearners, glmnet
#> * Predict Types:  [crank], distr, lp
#> * Feature Types: logical, integer, numeric
#> * Properties: selected_features, weights

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#> Measure: Partial Likelihood Deviance 
#> 
#>      Lambda Index Measure     SE Nonzero
#> min 0.00346    44   8.792 0.2481       6
#> 1se 0.05133    15   9.009 0.2321       3
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  28   9      0        180        1   107
#>   [2,]  35   5      1          2        0   172
#>   [3,]  35  10      1          9        0   106
#>   [4,]  34   5      0          5        0   120
#>   [5,]  37   9      0        180        1   151
#>   [6,]  36   1      0        180        1   155
#>   [7,]  35   0      0        180        1   119
#>   [8,]  38  12      1          8        1   120
#>   [9,]  33   6      1          1        1   115
#>  [10,]  38  16      1         10        0   160
#>  [11,]  42   3      1          1        1   130
#>  [12,]  38   5      1          3        0   125
#>  [13,]  40   6      0        180        1   138
#>  [14,]  42   2      0        180        0   100
#>  [15,]  43   3      1          0        1   100
#>  [16,]  40   1      1          0        1   145
#>  [17,]  43   4      1          0        1   130
#>  [18,]  42   4      0        180        0   162
#>  [19,]  40   3      1          1        0   170
#>  [20,]  42  12      1         10        1   170
#>  [21,]  43   2      1          1        1   116
#>  [22,]  42   2      0        180        1   124
#>  [23,]  44   5      1          1        0   170
#>  [24,]  45   3      0        180        1   154
#>  [25,]  44   3      0        180        0   141
#>  [26,]  41  13      1          1        0   140
#>  [27,]  45   6      0        180        1   170
#>  [28,]  41   5      1          4        1   141
#>  [29,]  44   2      1          1        1   150
#>  [30,]  43   2      0        180        1   140
#>  [31,]  46   2      1          1        0   126
#>  [32,]  47   4      1          3        0   118
#>  [33,]  48  15      0        180        1   160
#>  [34,]  45   3      0        150        0   130
#>  [35,]  44   3      1          0        1   180
#>  [36,]  46   7      1          2        0   166
#>  [37,]  45   4      1          0        0   124
#>  [38,]  43  10      0        180        0   185
#>  [39,]  47   6      1          0        1   116
#>  [40,]  46  13      1         10        0   100
#>  [41,]  44   0      1          0        1    96
#>  [42,]  47   4      1          3        1   160
#>  [43,]  43   3      1          0        1   124
#>  [44,]  45   8      1          0        1   117
#>  [45,]  49   5      0         73        1   136
#>  [46,]  46   6      1          0        1   100
#>  [47,]  44   4      1          0        1   114
#>  [48,]  47   2      0        180        0   108
#>  [49,]  44   9      1          8        1   135
#>  [50,]  45   5      0        180        1   190
#>  [51,]  44   2      0        180        0   142
#>  [52,]  46  15      0        180        1   120
#>  [53,]  45   9      1          0        1   145
#>  [54,]  47   3      1          1        1   120
#>  [55,]  48   3      0        180        0   154
#>  [56,]  46   3      1          0        1   119
#>  [57,]  49   4      0        180        0   117
#>  [58,]  47  10      0         10        1   140
#>  [59,]  50   1      1          0        1   129
#>  [60,]  47   7      0        180        0   145
#>  [61,]  50   4      1          1        0   125
#>  [62,]  50   6      1          2        1   140
#>  [63,]  49   7      1          7        1   110
#>  [64,]  46   9      1          9        1   122
#>  [65,]  49   2      0          2        0   105
#>  [66,]  51   1      0          1        1   145
#>  [67,]  47   2      0        180        0   150
#>  [68,]  49  23      0        179        1   112
#>  [69,]  46   6      1          0        1   156
#>  [70,]  52   2      0        180        1   170
#>  [71,]  50   7      1          0        1    92
#>  [72,]  50   4      0          4        1   100
#>  [73,]  50   9      0        180        0   130
#>  [74,]  49   7      1          4        1    90
#>  [75,]  47   8      0        180        0   160
#>  [76,]  47   6      0        180        1   162
#>  [77,]  52   2      0        180        0   155
#>  [78,]  46   3      0        180        1   120
#>  [79,]  46   1      1          1        0   145
#>  [80,]  50   4      1          1        0   150
#>  [81,]  53   8      0         36        1   160
#>  [82,]  47   2      1          1        0   110
#>  [83,]  52   4      1          4        0   152
#>  [84,]  49   9      1          3        0   102
#>  [85,]  49  15      0        180        1   160
#>  [86,]  53   5      0        180        1   140
#>  [87,]  54  17      1         12        1   102
#>  [88,]  53   5      0         77        0   159
#>  [89,]  53   7      1          0        0   199
#>  [90,]  51   3      1          1        0   140
#>  [91,]  49   5      1          2        1   150
#>  [92,]  52  14      1          7        1   200
#>  [93,]  48   6      0        180        0   160
#>  [94,]  53   4      1          0        1   156
#>  [95,]  51  13      0         99        1   160
#>  [96,]  54   9      1          0        1   138
#>  [97,]  49  16      0         16        0   125
#>  [98,]  55   3      1          1        0   150
#>  [99,]  55   6      1          2        1   114
#> [100,]  54   9      1          1        0   130
#> [101,]  55   4      1          2        0   150
#> [102,]  52   4      0        180        1   180
#> [103,]  54   4      1          0        1   121
#> [104,]  52   4      0        180        0   183
#> [105,]  55  28      1         13        1   160
#> [106,]  49   6      1          0        1   130
#> [107,]  50   7      1          1        0   156
#> [108,]  53   9      0          9        1    95
#> [109,]  53   8      1          0        1   130
#> [110,]  50   7      1          0        1   127
#> [111,]  56   4      1          1        1   130
#> [112,]  52   5      0        175        1   117
#> [113,]  55   2      0          2        0   145
#> [114,]  54   1      0        180        0   162
#> [115,]  54   7      1          0        1   100
#> [116,]  56   3      0        180        1   193
#> [117,]  56   2      0        180        0   132
#> [118,]  55   5      1          4        1   120
#> [119,]  52   8      0        180        0   119
#> [120,]  54   3      0        180        1   180
#> [121,]  55   6      0        180        0   170
#> [122,]  52  16      0         16        0   152
#> [123,]  53  10      1          9        0   172
#> [124,]  52  16      1         14        0   170
#> [125,]  53  15      0         15        1    90
#> [126,]  55   6      0        180        1   100
#> [127,]  55   6      1          5        1   138
#> [128,]  55   2      0        134        1   140
#> [129,]  56   3      0          8        1   139
#> [130,]  57   3      0          3        0   120
#> [131,]  54   7      1          2        0   129
#> [132,]  54   2      1          1        0   135
#> [133,]  52   9      1          3        0   170
#> [134,]  57   1      0        180        1   156
#> [135,]  57   1      0          1        1   100
#> [136,]  52   2      0        180        0   140
#> [137,]  55  11      1          7        0   104
#> [138,]  53   3      1          0        1   200
#> [139,]  57  10      0        180        1   170
#> [140,]  58   8      0          8        1   130
#> [141,]  55   3      1          1        1   156
#> [142,]  53  21      1         13        1   130
#> [143,]  59   3      1          1        0   172
#> [144,]  57   4      0        180        1   119
#> [145,]  58   6      1          0        1    90
#> [146,]  53  15      1         10        1   130
#> [147,]  55  13      0        166        1   140
#> [148,]  56   5      0          5        1   150
#> [149,]  53   4      0        147        1   145
#> [150,]  53   7      1          0        1   120
#> [151,]  57  11      1         10        1   129
#> [152,]  55   3      1          2        0   140
#> [153,]  55   5      0          5        1   131
#> [154,]  56   4      0          4        0   164
#> [155,]  58   9      1          0        1   180
#> [156,]  58   1      1          1        1   200
#> [157,]  55   2      0          2        0   106
#> [158,]  59   9      1          1        1   125
#> [159,]  60  11      1          9        0   106
#> [160,]  59   3      0        180        0   120
#> [161,]  58   4      1          0        1   160
#> [162,]  57   2      0          2        1   120
#> [163,]  60   5      1          1        0   138
#> [164,]  58  11      1          9        1   124
#> [165,]  55   5      1          0        1   160
#> [166,]  57  10      1          9        0   103
#> [167,]  59   6      1          0        1   140
#> [168,]  59   5      0        180        1   155
#> [169,]  58   4      1          3        0   120
#> [170,]  60   0      1          0        1    80
#> [171,]  59   2      1          1        0   140
#> [172,]  61   4      1          3        0   151
#> [173,]  61   9      1          8        0   150
#> [174,]  61   3      1          2        1   102
#> [175,]  58   1      0          1        1   100
#> [176,]  61  20      1         13        0   130
#> [177,]  58  10      0         10        1   150
#> [178,]  57   4      1          3        0   138
#> [179,]  57  11      0        180        1   150
#> [180,]  56  14      0         45        0   130
#> [181,]  56  18      1         11        1   165
#> [182,]  55   4      1          3        1   160
#> [183,]  58  11      0        172        1   135
#> [184,]  60  12      1          0        1   114
#> [185,]  61   4      1          0        1   115
#> [186,]  56   8      1          8        0   120
#> [187,]  61  13      1         12        1   130
#> [188,]  59  11      1          8        1   190
#> [189,]  57   1      0          1        0   126
#> [190,]  58   5      1          1        1   135
#> [191,]  61   8      0         77        0   120
#> [192,]  61  13      0         13        0   210
#> [193,]  58   8      1          5        0   152
#> [194,]  62  10      1          0        1   153
#> [195,]  62   7      1          2        1   180
#> [196,]  57   3      1          0        0   100
#> [197,]  61  18      0        170        0   140
#> [198,]  58   8      1          3        1   150
#> [199,]  57   7      0        169        0   180
#> [200,]  61   7      0          7        1   150
#> [201,]  59  13      1          2        0   198
#> [202,]  57  12      1          9        1   120
#> [203,]  62   4      1          0        0   160
#> [204,]  58   3      1          0        1   146
#> [205,]  62   4      1          3        0   173
#> [206,]  59   1      0        180        0   155
#> [207,]  59  16      1          9        1   133
#> [208,]  57  18      1          9        1    93
#> [209,]  61   5      0          5        1   160
#> [210,]  58  11      1          9        0   179
#> [211,]  62  17      1         10        1   180
#> [212,]  62   1      1          0        1   172
#> [213,]  63   3      1          1        0   180
#> [214,]  63   1      0        180        1   130
#> [215,]  61   7      0        180        0   135
#> [216,]  62   3      0        180        1   105
#> [217,]  63  15      1         10        1   126
#> [218,]  63   4      1          1        0   155
#> [219,]  60  18      1         13        0   132
#> [220,]  59   8      0        180        1   140
#> [221,]  61   9      1          9        1   150
#> [222,]  58   9      1          9        0   110
#> [223,]  62   7      0          7        0   150
#> [224,]  59   1      0         22        1   162
#> [225,]  58   2      0        180        0   127
#> [226,]  60   7      1          5        1   141
#> [227,]  60   7      0          7        0   140
#> [228,]  59   5      1          1        0   148
#> [229,]  60   7      1          1        1    90
#> [230,]  63   1      0          1        0   162
#> [231,]  60   3      0          3        0   168
#> [232,]  64  10      1          9        0   160
#> [233,]  62   6      0          6        0   120
#> [234,]  61   6      1          1        1   117
#> [235,]  64   6      1          0        1   140
#> [236,]  63  14      1          9        0   123
#> [237,]  65  36      1         11        0   140
#> [238,]  63   4      1          3        0   162
#> [239,]  66   3      1          1        0   127
#> [240,]  61  10      1          2        1   194
#> [241,]  64  32      1          9        1   160
#> [242,]  63   7      0        180        0   120
#> [243,]  66   5      1          0        1   110
#> [244,]  65   8      1          0        0   168
#> [245,]  64   0      0          0        1    90
#> [246,]  60   6      0        180        0   130
#> [247,]  64  21      1         10        0   190
#> [248,]  64   7      0        180        1   120
#> [249,]  66   6      1          1        1   130
#> [250,]  62   3      1          1        1   199
#> [251,]  65   6      0          9        0   112
#> [252,]  65   3      1          0        1    80
#> [253,]  63   5      1          4        0   170
#> [254,]  63   2      1          1        0   180
#> [255,]  62  13      1         11        0   180
#> [256,]  67  11      0         11        1   100
#> [257,]  66  16      1         11        1   169
#> [258,]  61  14      1          5        0   140
#> [259,]  63   3      1          2        0   120
#> [260,]  63   2      1          0        0   140
#> [261,]  64  19      1          8        1   160
#> [262,]  65   8      1          0        1   140
#> [263,]  67   6      0        180        1   170
#> [264,]  65  15      1         11        1   160
#> [265,]  68   5      1          4        1   150
#> [266,]  64  13      1         12        1   150
#> [267,]  64   6      1          0        1   125
#> [268,]  66   7      1          0        1   115
#> [269,]  66  13      1          0        0   118
#> [270,]  64  14      1         13        1   150
#> [271,]  65   3      0          3        0   105
#> [272,]  67   4      1          3        0   130
#> [273,]  65   2      1          1        1   170
#> [274,]  68   1      0        180        1   166
#> [275,]  68   5      0          5        1    90
#> [276,]  63  10      0         16        1   160
#> [277,]  66  14      0        180        0   130
#> [278,]  64   1      0          1        1   120
#> [279,]  68  18      0        180        1   260
#> [280,]  65  17      1         14        1   100
#> [281,]  63   8      1          1        1   162
#> [282,]  65  18      1          3        0   120
#> [283,]  63   1      1          0        1   155
#> [284,]  63  10      0         18        1   130
#> [285,]  68  11      0        180        0   160
#> [286,]  65  15      1         12        1   150
#> [287,]  66  11      1          0        0   100
#> [288,]  65   4      1          2        1   145
#> [289,]  69  12      0         15        1   140
#> [290,]  66  15      1         13        1   160
#> [291,]  63   2      0        180        0   150
#> [292,]  65  11      1          6        0   130
#> [293,]  69  21      1         10        0   180
#> [294,]  69   6      0        180        1   100
#> [295,]  68  14      1         13        1   140
#> [296,]  66   3      0          3        1   138
#> [297,]  69   1      1          0        0   170
#> [298,]  67   1      0        180        1   160
#> [299,]  68  10      1         10        1   150
#> [300,]  65   1      1          0        0   133
#> [301,]  63   2      1          0        0    99
#> [302,]  67   2      0        180        0   184
#> [303,]  65  10      1          1        1   148
#> [304,]  66  19      1         12        1   150
#> [305,]  65   4      1          1        0   130
#> [306,]  64   4      0        179        0   160
#> [307,]  70  15      1         12        1   132
#> [308,]  64   4      0        180        1   140
#> [309,]  66   7      1          5        1   131
#> [310,]  69   4      1          3        1   150
#> [311,]  69   8      0         93        0   140
#> [312,]  64  21      0         21        1   155
#> [313,]  65   1      0          1        1   120
#> [314,]  65   6      0        101        1   115
#> [315,]  70   7      1          0        1   190
#> [316,]  67   2      0        180        0   128
#> [317,]  66   9      1          3        1   151
#> [318,]  66   1      1          1        1   165
#> [319,]  70   4      1          0        1   180
#> [320,]  69   8      0        180        1   153
#> [321,]  70  14      0        171        0   166
#> [322,]  66   4      0        180        0   130
#> [323,]  67  10      1          9        0   200
#> [324,]  68  18      1         14        1   170
#> [325,]  69   0      0          0        1   148
#> [326,]  65   2      0        180        0   130
#> [327,]  68   7      1          0        1   150
#> [328,]  69   3      1          2        0   151
#> [329,]  67  14      1         13        0   130
#> [330,]  69   8      0        180        1   180
#> [331,]  66   2      0          2        1   228
#> [332,]  71   6      0         45        1   158
#> [333,]  69   5      0          5        1   142
#> [334,]  71   3      0        103        0   133
#> [335,]  69   3      0          3        1   130
#> [336,]  67   1      0         36        1   104
#> [337,]  67   5      0          5        0   130
#> [338,]  68   6      0        180        0   145
#> [339,]  69   8      1          5        1   195
#> [340,]  69   6      1          4        1   174
#> [341,]  72   7      0          7        1   110
#> [342,]  69   8      1          7        1   108
#> [343,]  67   3      0        180        0   110
#> [344,]  66   2      1          1        0   123
#> [345,]  69  19      0        180        0   130
#> [346,]  69  11      1          0        1   120
#> [347,]  66   2      0        180        0   130
#> [348,]  67   7      1          4        0   122
#> [349,]  69   4      1          3        0   132
#> [350,]  67  13      1          9        0   130
#> [351,]  70   3      0        123        0   130
#> [352,]  70   9      0        180        1   142
#> [353,]  72   5      1          4        0   170
#> [354,]  68   3      0         19        0   135
#> [355,]  67  12      1          8        0   120
#> [356,]  69   1      0          1        1   110
#> [357,]  69   5      0         76        0   120
#> [358,]  67   8      1          0        1   130
#> [359,]  72  13      1         11        1   195
#> [360,]  68  10      1          8        1   160
#> [361,]  70   7      0          7        0   102
#> [362,]  73  20      1          0        1   170
#> [363,]  71   6      0          9        0   120
#> [364,]  69  10      1          6        1   120
#> [365,]  70  11      0        180        1   210
#> [366,]  72  12      1         10        0   170
#> [367,]  67   8      0        180        1   170
#> [368,]  67   5      1          0        1   147
#> [369,]  67   9      0        180        0   158
#> [370,]  67   4      1          1        0   134
#> [371,]  72   6      1          5        0   115
#> [372,]  71   1      0        173        1   188
#> [373,]  68  23      0        180        1   220
#> [374,]  70   3      0        180        0   121
#> [375,]  69   3      0        180        0   220
#> [376,]  72   5      0         28        0   120
#> [377,]  73   6      0        180        1   117
#> [378,]  72  16      1          1        1   130
#> [379,]  69   1      1          0        0   155
#> [380,]  73   6      1          0        1   270
#> [381,]  71   2      1          0        1   180
#> [382,]  70   3      0          3        1   159
#> [383,]  70  13      1          9        0   100
#> [384,]  73   0      0        180        1   161
#> [385,]  73   4      0        180        1   154
#> [386,]  69   2      1          0        1   110
#> [387,]  71   3      1          1        0   150
#> [388,]  74  20      0         20        1   180
#> [389,]  71  20      1         10        0   140
#> [390,]  74   0      1          0        1    90
#> [391,]  73   3      1          0        1   136
#> [392,]  71  17      1         11        0   160
#> [393,]  71   8      1          7        0   149
#> [394,]  71   3      1          2        1   190
#> [395,]  70  26      1         11        1   120
#> [396,]  74   4      0          4        0   120
#> [397,]  73   4      0         58        1   160
#> [398,]  72   5      1          3        1   160
#> [399,]  70   3      0        180        1   154
#> [400,]  71   7      1          2        0   143
#> [401,]  74   3      0          3        1   150
#> [402,]  69   2      1          1        1    80
#> [403,]  70   4      1          0        1   140
#> [404,]  71  14      1         13        1   170
#> [405,]  74   7      1          0        1   117
#> [406,]  72  10      1          8        1   153
#> [407,]  69   7      0        180        1   144
#> [408,]  72  15      1         13        0   156
#> [409,]  70   8      0          8        0   120
#> [410,]  75   1      0          1        0   133
#> [411,]  75   2      1          1        0   145
#> [412,]  73  10      1         10        1   120
#> [413,]  74  15      1          9        1   179
#> [414,]  71   2      0         10        1   112
#> [415,]  73   1      0          1        1    80
#> [416,]  75   9      1          7        0   140
#> [417,]  75  13      1          1        1   130
#> [418,]  71  11      1          8        0   110
#> [419,]  72   1      1          1        0   168
#> [420,]  72   7      0         57        1   145
#> [421,]  73  10      0        180        0   162
#> [422,]  72  11      0         11        1   140
#> [423,]  70   3      0          3        0   150
#> [424,]  76  25      1         12        1   170
#> [425,]  73  12      1         12        1   140
#> [426,]  72   2      0        180        0   120
#> [427,]  75   1      0        180        1   140
#> [428,]  72   4      1          0        1   197
#> [429,]  71   3      1          0        0   144
#> [430,]  73   5      0        180        0   126
#> [431,]  73   4      0        180        0   124
#> [432,]  76   3      1          0        1   120
#> [433,]  71  32      1         12        1   107
#> [434,]  72   3      0        180        0   160
#> [435,]  77  11      0         11        1   150
#> [436,]  77   4      0          4        0   185
#> [437,]  72   7      1          2        0   142
#> [438,]  73  15      0         15        1   160
#> [439,]  71  16      0        180        0   140
#> [440,]  74   7      0        180        1   150
#> [441,]  74   3      0          3        1   128
#> [442,]  76   1      0        180        0   114
#> [443,]  74   2      1          1        0   140
#> [444,]  74  19      1          4        1   200
#> [445,]  75  23      1         14        1   110
#> [446,]  74   2      0        180        0   190
#> [447,]  76  17      1          0        1   200
#> [448,]  73   4      1          3        1   125
#> [449,]  76  13      1         10        0   110
#> [450,]  75   7      0          7        0   190
#> [451,]  75   0      0          0        1   130
#> [452,]  73  13      1         11        0   195
#> [453,]  75  12      0         12        1   160
#> [454,]  75   4      1          2        1   188
#> [455,]  74   6      0        180        0   160
#> [456,]  76   4      0          4        1   155
#> [457,]  75   1      0          1        1   125
#> [458,]  74   2      0        180        0   111
#> [459,]  73   0      0        180        0   156
#> [460,]  72   5      0        180        0   120
#> [461,]  76  44      1         10        0   105
#> [462,]  74  10      1          0        1   135
#> [463,]  74   8      1          8        1   170
#> [464,]  75   9      0        180        1   140
#> [465,]  73  33      1         12        1   175
#> [466,]  77   5      1          0        0   123
#> [467,]  77  12      1          9        1   100
#> [468,]  73  10      1          9        0   146
#> [469,]  77   1      1          0        1    90
#> [470,]  76  12      1         11        1   120
#> [471,]  78   5      1          0        1   170
#> [472,]  73   7      1          0        0   174
#> [473,]  74   6      0         79        1   140
#> [474,]  75   6      0        180        0   150
#> [475,]  79  10      1          8        0   190
#> [476,]  74   2      1          0        1   130
#> [477,]  78  18      0         18        1   144
#> [478,]  77   3      0        180        0   110
#> [479,]  73   8      1          1        1   162
#> [480,]  73  11      1          2        1   110
#> [481,]  78   7      0          7        1   133
#> [482,]  76  13      1          1        1   170
#> [483,]  80  10      1          6        1   147
#> [484,]  78   0      0        180        1   212
#> [485,]  78  13      1          5        0   130
#> [486,]  75  12      1          1        1   120
#> [487,]  78  15      0        180        1   270
#> [488,]  75  13      1          6        0   150
#> [489,]  74  10      1          8        0   135
#> [490,]  76   1      0          1        1    83
#> [491,]  79   4      0         80        0   145
#> [492,]  78  12      1          9        0   150
#> [493,]  78   2      1          1        0   130
#> [494,]  78  10      0        180        1   130
#> [495,]  76  11      1          0        0   120
#> [496,]  75  11      1          4        0   162
#> [497,]  77  24      0         24        1   160
#> [498,]  79   8      0         32        1   120
#> [499,]  80   9      0         23        1   128
#> [500,]  80   6      0          6        1   150
#> [501,]  78   6      1          0        1   240
#> [502,]  76   3      1          0        1   140
#> [503,]  78  14      1          0        1   140
#> [504,]  81   1      0          1        0   130
#> [505,]  78  11      1          8        1   118
#> [506,]  76   4      0          4        1   160
#> [507,]  79   4      0          4        1   125
#> [508,]  76  10      1          8        0   180
#> [509,]  76  12      1         10        1   127
#> [510,]  75   2      1          1        1   204
#> [511,]  78  11      0        180        1   135
#> [512,]  76   1      0          1        1   140
#> [513,]  77  31      1          3        1   161
#> [514,]  76   1      0          1        1    90
#> [515,]  79   4      1          0        1   120
#> [516,]  81   1      0        180        0   120
#> [517,]  80  15      1         12        1   150
#> [518,]  80  40      1          0        1   138
#> [519,]  78   4      0         59        1   112
#> [520,]  80  17      1         12        0   100
#> [521,]  76   7      0        161        0   151
#> [522,]  79  10      0         10        1   120
#> [523,]  81   4      1          2        1   126
#> [524,]  79  28      0        164        0   100
#> [525,]  80   9      0        118        1   186
#> [526,]  79   1      0         37        1   140
#> [527,]  81   3      0        180        0   184
#> [528,]  81   2      0        175        0   172
#> [529,]  78  26      1          5        0   194
#> [530,]  76   1      0        166        0   131
#> [531,]  81   4      1          1        1   104
#> [532,]  80   1      0          1        0   100
#> [533,]  77  10      1          8        1   130
#> [534,]  82   3      1          1        1   144
#> [535,]  77   5      0         85        0   188
#> [536,]  80   2      1          1        0   168
#> [537,]  79   6      0          6        0   152
#> [538,]  80   6      1          0        1   119
#> [539,]  80   5      0          5        1   130
#> [540,]  77   4      0        180        1    98
#> [541,]  81   1      0        108        0   129
#> [542,]  78  12      0        180        0   134
#> [543,]  79   1      0        125        0   193
#> [544,]  82  21      1          2        0   155
#> [545,]  84  22      1         10        0   180
#> [546,]  79   4      0          4        1   121
#> [547,]  80   6      0          6        1   110
#> [548,]  79   7      1          6        0   130
#> [549,]  83   4      0        103        0    97
#> [550,]  81  11      1          8        0   160
#> [551,]  81   5      0        177        0    41
#> [552,]  80  11      1          8        0   170
#> [553,]  78  23      1         10        1   145
#> [554,]  78   9      1          4        1   120
#> [555,]  82   8      1          1        0   128
#> [556,]  79   1      0        180        1   170
#> [557,]  84   5      1          1        1    85
#> [558,]  81  20      1          9        0   170
#> [559,]  83   8      0          8        0   115
#> [560,]  81  16      0         16        1   110
#> [561,]  80   6      1          0        1   150
#> [562,]  81   8      0        180        0   146
#> [563,]  80   8      1          7        0   160
#> [564,]  81   2      1          1        0   198
#> [565,]  82   6      0        128        1   100
#> [566,]  80   3      1          1        1   230
#> [567,]  84   1      0         38        1   205
#> [568,]  79   9      1          8        0   150
#> [569,]  85   3      1          2        1   160
#> [570,]  80   2      1          0        1   130
#> [571,]  79   4      0          4        1    60
#> [572,]  80   6      0         71        1   189
#> [573,]  82  19      0         19        0   120
#> [574,]  80  30      1         13        0   220
#> [575,]  83   9      0        180        0   198
#> [576,]  79  14      1          0        0   110
#> [577,]  81  14      1         12        1   128
#> [578,]  83   2      0        154        0   130
#> [579,]  83   1      0        180        0   160
#> [580,]  84  15      1         13        1   110
#> [581,]  81   1      0          1        1   145
#> [582,]  82   5      1          0        1   146
#> [583,]  81   4      0          4        0   160
#> [584,]  83  12      1          2        1   170
#> [585,]  81  19      1         14        0   120
#> [586,]  82   3      1          2        0   130
#> [587,]  82  15      1          0        0   183
#> [588,]  80   2      0         88        0   135
#> [589,]  83   7      0        126        0   135
#> [590,]  86   8      0          8        1   132
#> [591,]  81  16      1          9        0   180
#> [592,]  84   6      0        165        0   145
#> [593,]  84   3      0        180        1   120
#> [594,]  81  13      0        180        0   152
#> [595,]  81   2      1          0        1   118
#> [596,]  81   4      0        180        0   160
#> [597,]  83   9      0        180        1   149
#> [598,]  82   1      0        180        1   193
#> [599,]  83   4      0          4        0   130
#> [600,]  87   2      0          5        1   137
#> [601,]  86  12      1          0        1   132
#> [602,]  86   6      1          0        1   140
#> [603,]  84  16      0         70        1   150
#> [604,]  84   3      1          2        0   125
#> [605,]  86   2      0        180        1   169
#> [606,]  88  14      1          3        1   130
#> [607,]  83  13      1         12        0   170
#> [608,]  84   7      1          2        0   148
#> [609,]  87   2      0        180        0   113
#> [610,]  84   9      0         92        1   110
#> [611,]  84   3      0        180        1   170
#> [612,]  86   4      0         38        1   122
#> [613,]  82   4      0          4        0   130
#> [614,]  86  13      0        177        0   163
#> [615,]  86   6      0          6        1   117
#> [616,]  84  13      0         62        1   100
#> [617,]  86   6      1          1        0   112
#> [618,]  88   4      0          4        0   100
#> [619,]  83  20      1          3        1   150
#> [620,]  83   9      0         65        1   150
#> [621,]  86   6      0         46        0   173
#> [622,]  88   3      0        115        0   110
#> [623,]  88   2      0        180        1    68
#> [624,]  87   8      0          8        1   157
#> [625,]  86  15      1          8        1   109
#> [626,]  89   4      0          4        1   153
#> [627,]  87   6      0        180        1   110
#> [628,]  84   8      0        180        1   119
#> [629,]  85   8      0          8        1   136
#> [630,]  87  29      0         29        1    97
#> [631,]  87  15      1          9        1   138
#> [632,]  84   0      0        180        1   136
#> [633,]  89  10      0         46        1   170
#> [634,]  88   1      0          1        0   135
#> [635,]  91   8      0          8        0   100
#> [636,]  91  10      0        145        0   135
#> [637,]  87   6      0        126        1   168
#> [638,]  86  10      0        180        1   137
#> [639,]  86   9      1          7        0   130
#> [640,]  90   4      1          0        0   121
#> [641,]  87   5      0         36        1   150
#> [642,]  88   3      1          2        0   159
#> [643,]  89   3      1          1        1   160
#> [644,]  92   1      0          1        1   167
#> [645,]  91   3      0         33        1   137
#> [646,]  88   5      0        158        0   100
#> [647,]  89  12      1          0        1   130
#> [648,]  89   2      0        168        0   118
#> [649,]  91   5      0        169        1   176
#> [650,]  89  52      0         52        1   130
#> [651,]  92   7      0          7        1   110
#> [652,]  89   4      0          4        1   159
#> [653,]  91   0      0          0        0     0
#> [654,]  90  18      0        180        0   188
#> [655,]  91   4      1          0        1   120
#> [656,]  90  19      1         11        1   129
#> [657,]  94   6      0         50        0    78
#> [658,]  90   1      0          1        1   118
#> [659,]  91   2      0          2        1   116
#> [660,]  93   8      0        179        1   110
#> [661,]  90  16      0         16        1   106
#> [662,]  90   3      0         67        0   162
#> [663,]  96   3      0         12        1    97
#> [664,]  94   3      0         26        1   144
#> [665,]  91  12      0         53        1   212
#> [666,]  91   7      0          7        0   135
#> [667,]  93   0      1          0        1   122
#> [668,]  92   2      0          2        0   112
#> [669,]  93   4      0        180        1   135
#> [670,]  96  15      1          0        1   140
#> 
#> $y
#>   [1] 180.0+   5.0+ 180.0+   5.0+ 180.0+ 180.0+ 180.0+  12.0  180.0+ 180.0+
#>  [11] 180.0+   5.0+ 180.0+ 180.0+   3.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [21]   2.0+ 180.0+ 155.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0+ 180.0+ 180.0+
#>  [31] 180.0+ 180.0+ 180.0+ 150.0  180.0+ 180.0+ 180.0+ 180.0+   6.0+ 180.0+
#>  [41] 180.0+ 180.0+ 180.0+ 180.0+  73.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [51] 180.0+ 180.0+ 177.0+ 180.0+ 180.0+ 180.0+ 180.0+  10.0+ 172.0+ 180.0+
#>  [61] 180.0+ 180.0+   7.0  180.0+   2.0    1.0  180.0+ 179.0+ 180.0+ 180.0+
#>  [71] 180.0+   4.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [81]  36.0  180.0+   4.0+ 180.0+ 180.0+ 180.0+ 180.0+  77.0  180.0+ 180.0+
#>  [91] 180.0+  85.0  180.0+ 166.0+  99.0  180.0+  16.0+ 180.0+   6.0+ 180.0+
#> [101] 180.0+ 180.0+ 180.0+ 180.0+  28.0    6.0+ 180.0+   9.0+ 180.0+ 180.0+
#> [111] 180.0+ 175.0+   2.0  180.0+   7.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [121] 180.0+  16.0+ 180.0+  16.0   15.0+ 180.0+ 180.0+ 134.0+   8.0    3.0+
#> [131] 180.0+ 180.0+ 180.0+ 180.0+   1.0  180.0+ 180.0+ 180.0+ 180.0+   8.0+
#> [141] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 166.0+   5.0+ 147.0+ 180.0+
#> [151] 180.0+ 180.0+   5.0+   4.0+   9.0+   1.0    2.0+ 180.0+ 180.0+ 180.0+
#> [161] 180.0+   2.0  180.0+ 180.0+ 180.0+ 180.0+  64.0  180.0+ 180.0+   0.5 
#> [171] 180.0+ 180.0+ 180.0+   3.0    1.0  180.0+  10.0+ 180.0+ 180.0+  45.0 
#> [181] 180.0+ 180.0+ 172.0+ 172.0+ 180.0+   8.0  180.0+ 180.0+   1.0+ 180.0+
#> [191]  77.0   13.0+   8.0+ 180.0+ 180.0+ 180.0+ 170.0  180.0+ 169.0    7.0 
#> [201] 180.0+ 180.0+ 180.0+   3.0+ 180.0+ 180.0+ 180.0+  18.0    5.0+ 180.0+
#> [211] 180.0+   1.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+ 180.0+
#> [221] 180.0+   9.0    7.0+  22.0  180.0+  84.0    7.0+ 180.0+ 180.0+   1.0 
#> [231]   3.0+ 167.0    6.0+ 180.0+ 180.0+  14.0+  36.0  180.0+   3.0+  88.0 
#> [241] 180.0+ 180.0+ 180.0+ 180.0+   0.5  180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [251]   9.0    3.0  180.0+ 180.0+ 180.0+  11.0+ 180.0+ 180.0+   3.0+   2.0+
#> [261] 103.0   15.0  180.0+ 180.0+   5.0+  13.0  180.0+ 179.0+ 166.0+  14.0+
#> [271]   3.0  180.0+ 175.0+ 180.0+   5.0   16.0  180.0+   1.0  180.0+ 180.0+
#> [281] 180.0+ 123.0+   1.0+  18.0  180.0+  15.0+ 180.0+   4.0+  15.0  180.0+
#> [291] 180.0+ 180.0+ 174.0+ 180.0+ 180.0+   3.0  175.0  180.0+  10.0  180.0+
#> [301] 180.0+ 180.0+ 180.0+  19.0+ 180.0+ 179.0+ 180.0+ 180.0+   7.0+ 152.0+
#> [311]  93.0   21.0+   1.0  101.0    7.0+ 180.0+ 180.0+   1.0  180.0+ 180.0+
#> [321] 171.0  180.0+ 174.0+ 180.0+   0.5  180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [331]   2.0   45.0    5.0+ 103.0    3.0+  36.0    5.0+ 180.0+ 180.0+  97.0 
#> [341]   7.0    8.0+ 180.0+   2.0+ 180.0+ 180.0+ 180.0+   7.0  180.0+  13.0+
#> [351] 123.0  180.0+ 180.0+  19.0  180.0+   1.0   76.0  180.0+ 132.0   10.0+
#> [361]   7.0+ 124.0    9.0  180.0+ 180.0+  12.0  180.0+ 180.0+ 180.0+  76.0 
#> [371] 180.0+ 173.0+ 180.0+ 180.0+ 180.0+  28.0  180.0+  16.0+ 180.0+   6.0 
#> [381] 180.0+   3.0+  13.0+ 180.0+ 180.0+   2.0    3.0+  20.0   20.0    0.5 
#> [391] 180.0+ 180.0+   8.0    3.0  180.0+   4.0+  58.0  180.0+ 180.0+ 180.0+
#> [401]   3.0    2.0  180.0+  14.0+ 180.0+  10.0+ 180.0+ 180.0+   8.0+   1.0 
#> [411] 180.0+  15.0  180.0+  10.0    1.0  180.0+  13.0  180.0+   1.0   57.0 
#> [421] 180.0+  11.0    3.0+ 180.0+  12.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [431] 180.0+ 180.0+ 177.0+ 180.0+  11.0+   4.0+   7.0   15.0+ 180.0+ 180.0+
#> [441]   3.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+  17.0+ 180.0+ 174.0+   7.0 
#> [451]   0.5  180.0+  12.0   46.0  180.0+   4.0    1.0  180.0+ 180.0+ 180.0+
#> [461] 180.0+ 180.0+   8.0  180.0+  33.0    5.0  180.0+ 180.0+   1.0   12.0 
#> [471] 180.0+   7.0+  79.0  180.0+ 180.0+ 176.0+  18.0  180.0+ 180.0+  11.0 
#> [481]   7.0  180.0+  10.0  180.0+ 172.0   12.0  180.0+ 180.0+ 180.0+   1.0 
#> [491]  80.0  180.0+ 180.0+ 180.0+  11.0  152.0+  24.0   32.0   23.0    6.0 
#> [501] 180.0+   3.0+ 180.0+   1.0   11.0    4.0    4.0   10.0+ 180.0+   2.0+
#> [511] 180.0+   1.0  171.0    1.0  138.0  180.0+ 180.0+  40.0   59.0   17.0 
#> [521] 161.0   10.0+  93.0  164.0  118.0   37.0  180.0+ 175.0+ 171.0+ 166.0+
#> [531]  71.0    1.0   10.0  180.0+  85.0   10.0    6.0+   6.0    5.0  180.0+
#> [541] 108.0  180.0+ 125.0  180.0+ 180.0+   4.0    6.0  180.0+ 103.0  180.0+
#> [551] 177.0+ 169.0   70.0  180.0+ 180.0+ 180.0+ 180.0+  20.0    8.0+  16.0 
#> [561] 180.0+ 180.0+ 180.0+ 180.0+ 128.0    3.0+  38.0  180.0+ 180.0+ 180.0+
#> [571]   4.0   71.0   19.0   30.0  180.0+ 180.0+ 180.0+ 154.0  180.0+ 180.0+
#> [581]   1.0    5.0+   4.0+  77.0  180.0+   3.0   83.0   88.0  126.0    8.0 
#> [591] 180.0+ 165.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0+   5.0 
#> [601] 180.0+   6.0   70.0  180.0+ 180.0+  14.0   13.0  180.0+ 180.0+  92.0 
#> [611] 180.0+  38.0    4.0  177.0    6.0+  62.0    6.0+   4.0+  20.0   65.0 
#> [621]  46.0  115.0  180.0+   8.0+ 180.0+   4.0  180.0+ 180.0+   8.0   29.0 
#> [631] 180.0+ 180.0+  46.0    1.0+   8.0  145.0  126.0  180.0+ 180.0+   4.0 
#> [641]  36.0   75.0    3.0+   1.0   33.0  158.0  180.0+ 168.0  169.0   52.0 
#> [651]   7.0    4.0    0.5  180.0+   4.0  180.0+  50.0    1.0+   2.0  179.0+
#> [661]  16.0   67.0   12.0   26.0   53.0    7.0+   0.5    2.0  180.0+  15.0+
#> 
#> $weights
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8508312