Cross-Validated GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::cv.glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.cv.glmnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.cv.glmnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.cv_glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

foldid	untyped	NULL		-

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nfolds	integer	10		\([3, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	TRUE	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	lambda.1se		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

standardize.response	logical	FALSE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.gaussian	character	-	covariance, naive	-

type.logistic	character	Newton	Newton, modified.Newton	-

type.measure	character	deviance	deviance, C	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.coxnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvCVGlmnet

Methods

Public methods

LearnerSurvCVGlmnet$new()

LearnerSurvCVGlmnet$selected_features()

LearnerSurvCVGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvCVGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvCVGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvCVGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.cv_glmnet")
print(learner)
#> 
#> ── <LearnerSurvCVGlmnet> (surv.cv_glmnet): Regularized Generalized Linear Model 
#> • Model: -
#> • Parameters: use_pred_offset=TRUE
#> • Packages: mlr3, mlr3proba, mlr3extralearners, and glmnet
#> • Predict Types: [crank], distr, and lp
#> • Feature Types: logical, integer, and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: offset, selected_features, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#> Measure: Partial Likelihood Deviance 
#> 
#>      Lambda Index Measure     SE Nonzero
#> min 0.00339    44   2.751 0.1467       6
#> 1se 0.06662    12   2.882 0.1376       4
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  32   5      1          0        1   121
#>   [2,]  35  10      1          9        0   106
#>   [3,]  35   2      0        180        0   121
#>   [4,]  35   2      1          1        1   112
#>   [5,]  38  13      1          0        1   161
#>   [6,]  35   0      0        180        1   119
#>   [7,]  36   5      1          0        1   115
#>   [8,]  33   6      1          1        1   115
#>   [9,]  38  16      1         10        0   160
#>  [10,]  40  12      1          9        0   153
#>  [11,]  42   3      1          1        1   130
#>  [12,]  37   1      1          0        1   146
#>  [13,]  40   2      1          1        1   148
#>  [14,]  42   2      0        180        1   100
#>  [15,]  38   5      1          3        0   125
#>  [16,]  42   2      0          2        0   140
#>  [17,]  43   4      1          0        1   130
#>  [18,]  42   4      0        180        0   162
#>  [19,]  42  15      1         13        1   125
#>  [20,]  40   3      1          1        0   170
#>  [21,]  42  12      1         10        1   170
#>  [22,]  42   2      0        180        1   124
#>  [23,]  44   5      1          1        0   170
#>  [24,]  45   3      0        180        1   154
#>  [25,]  41  10      1          8        0   150
#>  [26,]  44   3      0        180        0   141
#>  [27,]  45   9      1          7        0   110
#>  [28,]  41   5      1          4        1   141
#>  [29,]  45   2      0        180        1   140
#>  [30,]  47   4      1          3        0   118
#>  [31,]  48  15      0        180        1   160
#>  [32,]  46   7      1          2        0   166
#>  [33,]  43  29      0        180        1   180
#>  [34,]  45   4      1          0        0   124
#>  [35,]  43  10      0        180        0   185
#>  [36,]  46  13      1         10        0   100
#>  [37,]  44   0      1          0        1    96
#>  [38,]  47   4      1          3        1   160
#>  [39,]  43   3      1          0        1   124
#>  [40,]  45   5      0          5        0   141
#>  [41,]  46   6      1          0        1   100
#>  [42,]  44   4      1          0        1   114
#>  [43,]  47   2      0        180        0   108
#>  [44,]  45   5      0        180        1   190
#>  [45,]  46   5      1          3        0   130
#>  [46,]  46   4      0        180        1   121
#>  [47,]  45   9      1          0        1   145
#>  [48,]  47   3      1          1        1   120
#>  [49,]  48   3      0        180        0   154
#>  [50,]  48  12      1         11        0   200
#>  [51,]  47   5      1          3        1   130
#>  [52,]  46   3      1          0        1   119
#>  [53,]  47  10      0         10        1   140
#>  [54,]  50   1      1          0        1   129
#>  [55,]  48   2      1          0        0   184
#>  [56,]  50   4      1          1        0   125
#>  [57,]  50   6      1          2        1   140
#>  [58,]  49   7      1          7        1   110
#>  [59,]  50   7      0        180        1   110
#>  [60,]  49  15      1         11        1   160
#>  [61,]  47   2      0        180        0   150
#>  [62,]  50   7      1          0        1    92
#>  [63,]  50   1      1          0        0   150
#>  [64,]  49   7      1          4        1    90
#>  [65,]  47   8      0        180        0   160
#>  [66,]  51   8      0        180        1   140
#>  [67,]  52   2      0        180        0   155
#>  [68,]  48   7      1          0        1   110
#>  [69,]  48  17      1         10        0   111
#>  [70,]  52   4      1          4        0   152
#>  [71,]  49  15      0        180        1   160
#>  [72,]  53   5      0        180        1   140
#>  [73,]  54  17      1         12        1   102
#>  [74,]  53   5      0         77        0   159
#>  [75,]  54   6      1          3        0   129
#>  [76,]  51   3      1          1        0   140
#>  [77,]  50  10      1          6        0   122
#>  [78,]  48   3      1          2        0   150
#>  [79,]  51  25      1          1        0   202
#>  [80,]  49   5      1          2        1   150
#>  [81,]  53   4      0          4        0   140
#>  [82,]  52  14      1          7        1   200
#>  [83,]  48   6      0        180        0   160
#>  [84,]  51  13      0         99        1   160
#>  [85,]  49  16      0         16        0   125
#>  [86,]  54  23      1         10        0   131
#>  [87,]  52   7      1          2        0   154
#>  [88,]  55   6      1          2        1   114
#>  [89,]  54   9      1          1        0   130
#>  [90,]  51  13      1         11        0   145
#>  [91,]  50   5      1          4        1   150
#>  [92,]  52   4      0        180        0   183
#>  [93,]  50   3      0        174        1   153
#>  [94,]  55  28      1         13        1   160
#>  [95,]  49   6      1          0        1   130
#>  [96,]  53   9      0          9        1    95
#>  [97,]  50   7      1          0        1   127
#>  [98,]  55   1      0        180        0   127
#>  [99,]  55   2      0          2        0   145
#> [100,]  54   1      0        180        0   162
#> [101,]  56   3      0        180        1   193
#> [102,]  56   2      0        180        0   132
#> [103,]  52   8      0        180        0   119
#> [104,]  53  18      1          9        1   150
#> [105,]  54   3      0        180        1   180
#> [106,]  52  16      0         16        0   152
#> [107,]  53  10      1          9        0   172
#> [108,]  52  16      1         14        0   170
#> [109,]  53  15      0         15        1    90
#> [110,]  55   6      1          5        1   138
#> [111,]  56   3      0          8        1   139
#> [112,]  54   7      1          2        0   129
#> [113,]  54   2      1          1        0   135
#> [114,]  54   2      1          1        1   176
#> [115,]  57   5      1          3        1   138
#> [116,]  57   1      0        180        1   156
#> [117,]  57   1      0          1        1   100
#> [118,]  52   2      0        180        0   140
#> [119,]  52  15      1         14        0   130
#> [120,]  53   3      1          0        1   200
#> [121,]  57  10      0        180        1   170
#> [122,]  58   8      0          8        1   130
#> [123,]  55   3      1          1        1   156
#> [124,]  57   0      0          0        1   150
#> [125,]  59   3      1          1        0   172
#> [126,]  58   6      1          0        1    90
#> [127,]  53  15      1         10        1   130
#> [128,]  54  17      1          8        1   227
#> [129,]  55   9      1          2        1   147
#> [130,]  55  13      0        166        1   140
#> [131,]  56   5      0          5        1   150
#> [132,]  54  23      1          8        0   120
#> [133,]  57   4      1          2        1   185
#> [134,]  53   4      0        147        1   145
#> [135,]  55   3      1          2        0   140
#> [136,]  55   5      0          5        1   131
#> [137,]  56   4      0          4        0   164
#> [138,]  58   1      1          1        1   200
#> [139,]  55   5      1          0        0   140
#> [140,]  55   2      0          2        0   106
#> [141,]  59   9      1          1        1   125
#> [142,]  57   1      0        180        0   148
#> [143,]  57   2      0          2        1   120
#> [144,]  60   5      1          1        0   138
#> [145,]  58  11      1          9        1   124
#> [146,]  55   5      1          0        1   160
#> [147,]  57  10      1          9        0   103
#> [148,]  59   6      1          0        1   140
#> [149,]  59   5      0        180        1   155
#> [150,]  61   9      0          9        1   160
#> [151,]  60   0      1          0        1    80
#> [152,]  59   2      1          1        0   140
#> [153,]  58   8      0        161        1   140
#> [154,]  58  14      1          6        0   190
#> [155,]  61   4      1          3        0   151
#> [156,]  61   9      1          8        0   150
#> [157,]  61   3      1          2        1   102
#> [158,]  58   1      0          1        1   100
#> [159,]  61  20      1         13        0   130
#> [160,]  57   4      1          3        0   138
#> [161,]  57  11      0        180        1   150
#> [162,]  61   3      0         17        0   143
#> [163,]  56  14      0         45        0   130
#> [164,]  57   3      1          2        0   120
#> [165,]  58  19      1         13        1   140
#> [166,]  56  13      1          6        1   158
#> [167,]  56  18      1         11        1   165
#> [168,]  59   9      1          0        1    80
#> [169,]  55   4      1          3        1   160
#> [170,]  58  11      0        172        1   135
#> [171,]  55   9      1          7        1   135
#> [172,]  61   4      1          0        1   115
#> [173,]  61  13      1         12        1   130
#> [174,]  57   1      0          1        0   126
#> [175,]  57  15      1         13        1   110
#> [176,]  59  10      0        180        0   160
#> [177,]  61   8      0         77        0   120
#> [178,]  58   8      1          5        0   152
#> [179,]  62  10      1          0        1   153
#> [180,]  57   3      1          0        0   100
#> [181,]  61  28      1          7        0   133
#> [182,]  58   8      1          3        1   150
#> [183,]  57   7      0        169        0   180
#> [184,]  61   7      0          7        1   150
#> [185,]  60   7      0          7        0   147
#> [186,]  61   6      0          6        0   134
#> [187,]  59  13      1          2        0   198
#> [188,]  60  17      1          8        1   140
#> [189,]  58   3      1          0        1   146
#> [190,]  59  16      1          9        1   133
#> [191,]  63   6      0         28        1   120
#> [192,]  61  13      0         13        0   120
#> [193,]  61   5      0          5        1   110
#> [194,]  57  18      1          9        1    93
#> [195,]  61   5      0          5        1   160
#> [196,]  58  11      1          9        0   179
#> [197,]  57   2      1          1        0   159
#> [198,]  62   1      1          0        1   172
#> [199,]  63   1      0        180        1   130
#> [200,]  61   7      0        180        0   135
#> [201,]  63   4      1          3        0   222
#> [202,]  62   3      0        180        1   105
#> [203,]  63  15      1         10        1   126
#> [204,]  64   4      0        180        0   130
#> [205,]  63   4      1          1        0   155
#> [206,]  60  18      1         13        0   132
#> [207,]  61   9      1          9        1   150
#> [208,]  58   9      1          9        0   110
#> [209,]  58   2      0        180        0   127
#> [210,]  59   4      0        180        0   196
#> [211,]  60   7      0          7        0   140
#> [212,]  60   7      1          1        1    90
#> [213,]  65  13      0        180        1   100
#> [214,]  63   1      0          1        0   162
#> [215,]  63   1      0          1        0   130
#> [216,]  62   6      0        180        0   170
#> [217,]  61  15      1         13        0   170
#> [218,]  59   4      0          4        0   149
#> [219,]  60   3      0          3        0   168
#> [220,]  64  10      1          9        0   160
#> [221,]  62   6      0          6        0   120
#> [222,]  63  12      1         10        0   200
#> [223,]  59  10      0        180        1   130
#> [224,]  60   8      0         17        1   130
#> [225,]  61   6      1          1        1   117
#> [226,]  64  12      1         11        0   160
#> [227,]  64   6      1          0        1   140
#> [228,]  63  10      1          0        1   148
#> [229,]  65  36      1         11        0   140
#> [230,]  63   4      1          3        0   162
#> [231,]  61  10      1          2        1   194
#> [232,]  64  32      1          9        1   160
#> [233,]  63   7      0        180        0   120
#> [234,]  66   5      1          0        1   110
#> [235,]  65  10      1          8        1   120
#> [236,]  60   6      0        180        0   130
#> [237,]  64  21      1         10        0   190
#> [238,]  61  12      1         11        0   154
#> [239,]  61   4      0        180        1   113
#> [240,]  63  16      1          7        1   110
#> [241,]  66   6      1          1        1   130
#> [242,]  62   3      1          1        1   199
#> [243,]  65   6      0          9        0   112
#> [244,]  65   3      1          0        1    80
#> [245,]  62  13      1         11        0   180
#> [246,]  67  11      0         11        1   100
#> [247,]  64   2      0          2        0   201
#> [248,]  66  18      1          5        0   142
#> [249,]  66  16      1         11        1   169
#> [250,]  62   9      0        180        0   145
#> [251,]  61  14      1          5        0   140
#> [252,]  61  15      1         10        0   130
#> [253,]  63   9      1          8        1   160
#> [254,]  63   3      1          2        0   120
#> [255,]  63   2      1          0        0   140
#> [256,]  64  13      1         12        1   150
#> [257,]  64   6      1          0        1   125
#> [258,]  66   7      1          0        1   115
#> [259,]  64   0      0          0        1   148
#> [260,]  67   4      1          3        0   130
#> [261,]  66   3      1          0        1   135
#> [262,]  66   6      1          0        1   140
#> [263,]  65   2      1          1        1   170
#> [264,]  64  10      1          9        1   110
#> [265,]  63   7      1          0        0   162
#> [266,]  68   5      0          5        1    90
#> [267,]  63  10      0         16        1   160
#> [268,]  64   1      0          1        1   120
#> [269,]  68  18      0        180        1   260
#> [270,]  65  17      1         14        1   100
#> [271,]  63   8      1          1        1   162
#> [272,]  65  18      1          3        0   120
#> [273,]  63   1      1          0        1   155
#> [274,]  67  11      0         11        0   150
#> [275,]  68  14      0         79        0   172
#> [276,]  66  12      1         10        1   150
#> [277,]  66  11      1          0        0   100
#> [278,]  69  12      0         15        1   140
#> [279,]  66  15      1         13        1   160
#> [280,]  63   2      0        180        0   150
#> [281,]  65  11      1          6        0   130
#> [282,]  69  21      1         10        0   180
#> [283,]  69   6      0        180        1   100
#> [284,]  66   9      1          8        0   130
#> [285,]  65   8      1          0        1    90
#> [286,]  67   1      0        180        1   160
#> [287,]  65   1      1          0        0   133
#> [288,]  67   7      1          4        1   130
#> [289,]  67   2      0        180        0   184
#> [290,]  65  10      1          1        1   148
#> [291,]  66  19      1         12        1   150
#> [292,]  69   6      0         99        1   140
#> [293,]  65   4      1          1        0   130
#> [294,]  64   4      0        179        0   160
#> [295,]  66   4      0        180        1   130
#> [296,]  70  15      1         12        1   132
#> [297,]  64  11      0         11        0   125
#> [298,]  64   0      1          0        1   118
#> [299,]  67   2      0         18        0   131
#> [300,]  66   7      1          5        1   131
#> [301,]  66   4      0        180        0   177
#> [302,]  69   4      1          3        1   150
#> [303,]  65  13      1         12        1   130
#> [304,]  69   8      0         93        0   140
#> [305,]  66   6      0        180        0   140
#> [306,]  65   6      0        101        1   115
#> [307,]  68   4      0          4        1   190
#> [308,]  71   3      0          5        0   112
#> [309,]  70   7      1          0        1   190
#> [310,]  71  20      1          0        1   160
#> [311,]  67   2      0        180        0   128
#> [312,]  66   9      1          3        1   151
#> [313,]  66   1      1          1        1   165
#> [314,]  70   4      1          0        1   180
#> [315,]  69   8      0        180        1   153
#> [316,]  70  14      0        171        0   166
#> [317,]  66   4      0        180        0   130
#> [318,]  65   2      0        180        0   130
#> [319,]  68   7      1          0        1   150
#> [320,]  69   3      1          2        0   151
#> [321,]  67  14      1         13        0   130
#> [322,]  65  14      1         13        1   150
#> [323,]  69   8      0        180        1   180
#> [324,]  71   7      0          7        0   230
#> [325,]  66   2      0          2        1   228
#> [326,]  71   6      0         45        1   158
#> [327,]  69   5      0          5        1   142
#> [328,]  71   3      0        103        0   133
#> [329,]  70  22      1         13        0   103
#> [330,]  67   5      0          5        0   130
#> [331,]  69   8      1          5        1   195
#> [332,]  69   8      1          7        1   108
#> [333,]  67   3      0        180        0   110
#> [334,]  66   2      1          1        0   123
#> [335,]  68  18      0         18        1   100
#> [336,]  67  14      0        172        1   140
#> [337,]  69  11      1          0        1   120
#> [338,]  66   2      0        180        0   130
#> [339,]  67   7      1          4        0   122
#> [340,]  69   4      1          3        0   132
#> [341,]  68   2      0          7        1   130
#> [342,]  70   3      0        123        0   130
#> [343,]  70   9      0        180        1   142
#> [344,]  72   5      1          4        0   170
#> [345,]  67  22      1          1        1   140
#> [346,]  67  12      1          8        0   120
#> [347,]  67   1      0          1        1    60
#> [348,]  67   4      0         60        1   136
#> [349,]  69   5      0         76        0   120
#> [350,]  67   8      1          0        1   130
#> [351,]  70  35      1          0        1   105
#> [352,]  72  30      1          0        1   145
#> [353,]  73  20      1          0        1   170
#> [354,]  70  11      0        180        1   210
#> [355,]  72  19      1          8        0   120
#> [356,]  67   8      0        180        1   170
#> [357,]  67   5      1          0        1   147
#> [358,]  70   5      0        180        0   150
#> [359,]  67   4      1          1        0   134
#> [360,]  71   1      0        173        1   188
#> [361,]  70   3      0        180        0   121
#> [362,]  69   3      0        180        0   220
#> [363,]  71   3      1          2        0   150
#> [364,]  72   5      0         28        0   120
#> [365,]  73   6      0        180        1   117
#> [366,]  69  16      1         10        1   140
#> [367,]  72  16      1          1        1   130
#> [368,]  69   1      1          0        0   155
#> [369,]  71   2      1          0        1   180
#> [370,]  73   7      0          7        1   140
#> [371,]  70   3      0          3        1   159
#> [372,]  72   6      0        180        1   130
#> [373,]  73   0      0        180        1   161
#> [374,]  74   8      1          0        1    85
#> [375,]  73   4      0        180        1   154
#> [376,]  69   2      1          0        1   110
#> [377,]  71   3      1          1        0   150
#> [378,]  74  20      0         20        1   180
#> [379,]  71  20      1         10        0   140
#> [380,]  71   8      1          7        0   149
#> [381,]  69  12      1          1        1   149
#> [382,]  74   4      0          4        0   120
#> [383,]  73   4      0         58        1   160
#> [384,]  70   3      0        180        1   154
#> [385,]  71   7      1          2        0   143
#> [386,]  72   8      1          0        1   140
#> [387,]  74   3      0          3        1   150
#> [388,]  71  13      1          8        0   121
#> [389,]  70   4      1          0        1   140
#> [390,]  72  10      1          8        1   153
#> [391,]  72  15      1         13        0   156
#> [392,]  71  10      1          9        1   120
#> [393,]  75   1      0          1        0   133
#> [394,]  73  10      1          9        1   146
#> [395,]  73  10      1         10        1   120
#> [396,]  74  15      1          9        1   179
#> [397,]  71   2      0         10        1   112
#> [398,]  73   1      0          1        1    80
#> [399,]  75   9      1          7        0   140
#> [400,]  75  13      1          1        1   130
#> [401,]  71  11      1          8        0   110
#> [402,]  71   4      0          4        0   134
#> [403,]  73  10      1          8        0   120
#> [404,]  70   7      1          4        0   184
#> [405,]  72   1      1          1        0   168
#> [406,]  72   7      0         57        1   145
#> [407,]  73  10      0        180        0   162
#> [408,]  72  11      0         11        1   140
#> [409,]  70   3      0          3        0   150
#> [410,]  76  25      1         12        1   170
#> [411,]  75   1      0        180        1   140
#> [412,]  72   4      1          0        1   197
#> [413,]  71   3      1          0        0   144
#> [414,]  73   5      0        180        0   126
#> [415,]  73   4      0        180        0   124
#> [416,]  76   3      1          0        1   120
#> [417,]  71  32      1         12        1   107
#> [418,]  72   3      0        180        0   160
#> [419,]  76   5      0          5        1   130
#> [420,]  77  11      0         11        1   150
#> [421,]  77   4      0          4        0   185
#> [422,]  75   3      1          1        0   180
#> [423,]  74   3      0          3        1   128
#> [424,]  76   1      0        180        0   114
#> [425,]  76   8      1          0        1   141
#> [426,]  74  19      1          4        1   200
#> [427,]  73   6      0          6        1   114
#> [428,]  75  23      1         14        1   110
#> [429,]  72   4      0         85        1   120
#> [430,]  76  17      1          0        1   200
#> [431,]  73   4      1          3        1   125
#> [432,]  75   4      1          0        1   122
#> [433,]  75   7      0          7        0   190
#> [434,]  73  13      1         11        0   195
#> [435,]  75  12      0         12        1   160
#> [436,]  74   8      1          0        1   105
#> [437,]  76  13      1          8        1   148
#> [438,]  74   6      0        180        0   160
#> [439,]  75   1      0          1        1   125
#> [440,]  74   2      0        180        0   111
#> [441,]  76   5      0        180        0   185
#> [442,]  74  10      1          0        1   135
#> [443,]  76   5      1          0        1   167
#> [444,]  73  33      1         12        1   175
#> [445,]  77  12      1          9        1   100
#> [446,]  73  10      1          9        0   146
#> [447,]  77  12      0        180        0   130
#> [448,]  77   1      1          0        1    90
#> [449,]  78   5      1          0        1   170
#> [450,]  73   7      1          0        0   174
#> [451,]  74   6      0         79        1   140
#> [452,]  75   3      1          1        1   171
#> [453,]  75   6      0        180        0   150
#> [454,]  79  10      1          8        0   190
#> [455,]  74   2      1          0        1   130
#> [456,]  76  29      0         47        0    90
#> [457,]  73   8      1          1        1   162
#> [458,]  73  11      1          2        1   110
#> [459,]  74   2      0        180        0   100
#> [460,]  78   7      0          7        1   133
#> [461,]  74  15      0        180        1   172
#> [462,]  78   8      1          6        1   110
#> [463,]  74   7      0          7        0   161
#> [464,]  79   6      0        180        0   170
#> [465,]  80  10      1          6        1   147
#> [466,]  78   0      0        180        1   212
#> [467,]  78  13      1          5        0   130
#> [468,]  75   5      0        119        1   150
#> [469,]  75  12      1          1        1   120
#> [470,]  75  13      1          6        0   150
#> [471,]  74  10      1          8        0   135
#> [472,]  76   1      0          1        1    83
#> [473,]  79   4      0         80        0   145
#> [474,]  78   2      1          1        0   130
#> [475,]  75   4      1          0        0   212
#> [476,]  77   2      1          0        1   143
#> [477,]  75  11      1          4        0   162
#> [478,]  75   3      0          3        0     0
#> [479,]  76   7      0         29        1   150
#> [480,]  77  24      0         24        1   160
#> [481,]  79   8      0         32        1   120
#> [482,]  80   6      0          6        1   150
#> [483,]  78   6      1          0        1   240
#> [484,]  76   3      1          0        1   140
#> [485,]  78  11      1          1        1   140
#> [486,]  79  11      0        180        0   160
#> [487,]  79   2      1          0        1   121
#> [488,]  78  14      1          0        1   140
#> [489,]  78  11      1          8        1   118
#> [490,]  76   4      0          4        1   160
#> [491,]  76  10      1          8        0   180
#> [492,]  77   6      0          6        1   107
#> [493,]  75   2      1          1        1   204
#> [494,]  78  11      0        180        1   135
#> [495,]  76   1      0          1        1   140
#> [496,]  77  31      1          3        1   161
#> [497,]  76   1      0          1        1    90
#> [498,]  78   7      1          0        1   110
#> [499,]  79   3      0          3        0   120
#> [500,]  77   7      0        180        1   170
#> [501,]  77   6      0          6        1   144
#> [502,]  79   4      1          0        1   120
#> [503,]  81   1      0        180        0   120
#> [504,]  80  15      1         12        1   150
#> [505,]  77   9      1          4        0   141
#> [506,]  82   5      0          8        1   120
#> [507,]  80  40      1          0        1   138
#> [508,]  78   4      0         59        1   112
#> [509,]  76   7      0        161        0   151
#> [510,]  79  10      0         10        1   120
#> [511,]  81   4      1          2        1   126
#> [512,]  79  28      0        164        0   100
#> [513,]  80   9      0        118        1   186
#> [514,]  80   6      0        173        1   160
#> [515,]  78  32      0        180        1   130
#> [516,]  79   1      0         37        1   140
#> [517,]  81   3      0        180        0   184
#> [518,]  78   7      0          7        1   147
#> [519,]  80   5      1          1        1   108
#> [520,]  78   4      0        180        0   175
#> [521,]  79   3      0          3        1   101
#> [522,]  78  26      1          5        0   194
#> [523,]  76   1      0        166        0   131
#> [524,]  81   4      1          1        1   104
#> [525,]  80   1      0          1        0   100
#> [526,]  78   3      1          1        1   152
#> [527,]  77   5      0         85        0   188
#> [528,]  80   2      1          1        0   168
#> [529,]  80   6      1          0        1   119
#> [530,]  78   2      0        180        0   148
#> [531,]  82   1      1          0        1    82
#> [532,]  79  10      0        180        1   150
#> [533,]  77   4      0        180        1    98
#> [534,]  78  12      0        180        0   134
#> [535,]  79   1      0        125        0   193
#> [536,]  79   4      0          4        1   121
#> [537,]  80   6      0          6        1   110
#> [538,]  83   9      1          5        1   170
#> [539,]  82   5      0        180        0   110
#> [540,]  83   5      0        180        0   148
#> [541,]  79   7      1          6        0   130
#> [542,]  83   4      0        103        0    97
#> [543,]  81  11      1          8        0   160
#> [544,]  79   4      0          4        1   183
#> [545,]  78   9      1          4        1   120
#> [546,]  82   8      1          1        0   128
#> [547,]  79   1      0        180        1   170
#> [548,]  81  20      1          9        0   170
#> [549,]  81  16      0         16        1   110
#> [550,]  80   6      1          0        1   150
#> [551,]  81   8      0        180        0   146
#> [552,]  80   8      1          7        0   160
#> [553,]  79   7      0        177        0   197
#> [554,]  79   0      1          0        1    96
#> [555,]  85   4      0        180        0    90
#> [556,]  81   2      1          1        0   198
#> [557,]  83   2      0          2        1   155
#> [558,]  80   3      1          1        1   230
#> [559,]  82  23      1          0        0   110
#> [560,]  84   5      0        180        1   203
#> [561,]  84   4      0          4        1    85
#> [562,]  84   1      0         38        1   205
#> [563,]  83   3      0        180        0   174
#> [564,]  81   4      0         90        1   138
#> [565,]  79   9      1          8        0   150
#> [566,]  85   3      1          2        1   160
#> [567,]  84   4      0         89        1   129
#> [568,]  79   4      0          4        1    60
#> [569,]  80   6      0         71        1   189
#> [570,]  83   1      0          1        1   100
#> [571,]  82  19      0         19        0   120
#> [572,]  83   9      0        180        0   198
#> [573,]  83   3      0        114        0    98
#> [574,]  81  14      1         12        1   128
#> [575,]  83   1      0        180        0   160
#> [576,]  81   4      0          4        0   175
#> [577,]  84  15      1         13        1   110
#> [578,]  81   1      0          1        1   145
#> [579,]  81  12      0         12        1   163
#> [580,]  82  16      1          8        0   103
#> [581,]  81   4      0          4        0   160
#> [582,]  86  12      0        180        1   120
#> [583,]  82   3      1          2        0   130
#> [584,]  82  15      1          0        0   183
#> [585,]  80   2      0         88        0   135
#> [586,]  83   7      0        126        0   135
#> [587,]  81  16      1          9        0   180
#> [588,]  84   6      0        165        0   145
#> [589,]  86   3      0          3        1   140
#> [590,]  82   9      0        180        1   134
#> [591,]  84   3      0        180        1   120
#> [592,]  81  13      0        180        0   152
#> [593,]  85   3      0          3        1   118
#> [594,]  81   2      1          0        1   118
#> [595,]  81   4      0        180        0   160
#> [596,]  83   9      0        180        1   149
#> [597,]  82   1      0        180        1   193
#> [598,]  83   4      0          4        0   130
#> [599,]  86  12      1          0        1   132
#> [600,]  84  16      0         70        1   150
#> [601,]  83  19      0         43        0   150
#> [602,]  84   3      1          2        0   125
#> [603,]  86   2      0        180        1   169
#> [604,]  88  14      1          3        1   130
#> [605,]  83  13      1         12        0   170
#> [606,]  84   7      1          2        0   148
#> [607,]  87   2      0        180        0   113
#> [608,]  84   9      0         92        1   110
#> [609,]  84   3      0        180        1   170
#> [610,]  82   4      0          4        0   130
#> [611,]  85   3      0          3        1   113
#> [612,]  86   6      0          6        1   117
#> [613,]  84  13      0         62        1   100
#> [614,]  86   6      1          1        0   112
#> [615,]  88   4      0          4        0   100
#> [616,]  83   9      0         65        1   150
#> [617,]  86   9      1          7        1   142
#> [618,]  87   2      0        180        1   130
#> [619,]  86   6      0         46        0   173
#> [620,]  83   3      0          3        1   130
#> [621,]  87   8      0          8        1   157
#> [622,]  88   4      0          4        0    86
#> [623,]  89   5      0        119        1   140
#> [624,]  87   6      0        180        1   110
#> [625,]  87   1      0          1        0   170
#> [626,]  84   8      0        180        1   119
#> [627,]  85   8      0          8        1   136
#> [628,]  84   2      0        110        1   174
#> [629,]  87  15      1          9        1   138
#> [630,]  84   0      0        180        1   136
#> [631,]  88   1      0          1        0   135
#> [632,]  86   4      0        180        1   145
#> [633,]  91   8      0          8        0   100
#> [634,]  87   6      1          0        0   125
#> [635,]  88   7      0         24        0   119
#> [636,]  88   8      0         50        1   154
#> [637,]  87   6      0        126        1   168
#> [638,]  86  10      0        180        1   137
#> [639,]  86   9      1          7        0   130
#> [640,]  90   4      1          0        0   121
#> [641,]  87  43      0        178        1   130
#> [642,]  87   5      0         36        1   150
#> [643,]  90   5      1          0        1   125
#> [644,]  88   3      1          2        0   159
#> [645,]  92   1      0          1        1   167
#> [646,]  91   3      0         33        1   137
#> [647,]  88   5      0        158        0   100
#> [648,]  87   7      0         74        1   105
#> [649,]  89  12      1          0        1   130
#> [650,]  89   2      0        168        0   118
#> [651,]  89  52      0         52        1   130
#> [652,]  92   7      0          7        1   110
#> [653,]  89   4      0          4        1   159
#> [654,]  90  18      0        180        0   188
#> [655,]  91   4      1          0        1   120
#> [656,]  94   6      0         50        0    78
#> [657,]  90   1      0          1        1   118
#> [658,]  91   2      0          2        1   116
#> [659,]  93   8      0        179        1   110
#> [660,]  94   8      0          8        1   142
#> [661,]  92   4      0         76        1   149
#> [662,]  90  16      0         16        1   106
#> [663,]  90   3      0         67        0   162
#> [664,]  95   8      1          5        1   150
#> [665,]  94   3      0         26        1   144
#> [666,]  91  12      0         53        1   212
#> [667,]  91   7      0          7        0   135
#> [668,]  92   5      0         69        0   139
#> [669,]  93   4      0        180        1   135
#> [670,]  96   3      1          0        1   104
#> 
#> $y
#>   [1]   5.0+ 180.0+ 180.0+   2.0+ 180.0+ 180.0+   5.0+ 180.0+ 180.0+ 180.0+
#>  [11] 180.0+ 180.0+   2.0+ 180.0+   5.0+   2.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [21] 180.0+ 180.0+ 155.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0+ 180.0+ 180.0+
#>  [31] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0+
#>  [41] 180.0+ 180.0+ 180.0+ 180.0+   5.0+ 180.0+ 177.0+ 180.0+ 180.0+ 180.0+
#>  [51] 180.0+ 180.0+  10.0+ 172.0+ 180.0+ 180.0+ 180.0+   7.0  180.0+ 179.0+
#>  [61] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   7.0   88.0+   4.0+
#>  [71] 180.0+ 180.0+ 180.0+  77.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [81]   4.0+  85.0  180.0+  99.0   16.0+ 152.0+   7.0+   6.0+ 180.0+  13.0+
#>  [91] 171.0+ 180.0+ 174.0+  28.0    6.0+   9.0+ 180.0+ 180.0+   2.0  180.0+
#> [101] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  16.0+ 180.0+  16.0   15.0+ 180.0+
#> [111]   8.0  180.0+ 180.0+ 180.0+ 140.0  180.0+   1.0  180.0+ 180.0+ 180.0+
#> [121] 180.0+   8.0+ 180.0+   0.5  180.0+ 180.0+ 180.0+ 171.0+  15.0  166.0+
#> [131]   5.0+ 180.0+   4.0+ 147.0+ 180.0+   5.0+   4.0+   1.0  180.0+   2.0+
#> [141] 180.0+ 180.0+   2.0  180.0+ 180.0+ 180.0+ 180.0+  64.0  180.0+   9.0+
#> [151]   0.5  180.0+ 161.0+ 171.0+ 180.0+ 180.0+   3.0    1.0  180.0+ 180.0+
#> [161] 180.0+  17.0   45.0    3.0+  19.0  180.0+ 180.0+   9.0+ 180.0+ 172.0+
#> [171]  24.0  180.0+ 180.0+   1.0+  15.0  180.0+  77.0    8.0+ 180.0+ 180.0+
#> [181]  94.0  180.0+ 169.0    7.0    7.0+   6.0  180.0+ 180.0+   3.0+ 180.0+
#> [191]  28.0   13.0+   5.0   18.0    5.0+ 180.0+ 180.0+   1.0  180.0+ 180.0+
#> [201] 180.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+ 180.0+   9.0  180.0+ 180.0+
#> [211]   7.0+ 180.0+ 180.0+   1.0    1.0  180.0+ 180.0+   4.0+   3.0+ 167.0 
#> [221]   6.0+ 180.0+ 180.0+  17.0  180.0+  12.0  180.0+ 180.0+  36.0  180.0+
#> [231]  88.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  12.0+ 180.0+ 180.0+
#> [241] 180.0+ 180.0+   9.0    3.0  180.0+  11.0+   2.0+  18.0+ 180.0+ 180.0+
#> [251] 180.0+ 180.0+ 180.0+   3.0+   2.0+  13.0  180.0+ 179.0+   0.5+ 180.0+
#> [261]   3.0+ 180.0+ 175.0+ 180.0+   7.0+   5.0   16.0    1.0  180.0+ 180.0+
#> [271] 180.0+ 123.0+   1.0+  11.0+  79.0   80.0  180.0+  15.0  180.0+ 180.0+
#> [281] 180.0+ 174.0+ 180.0+ 180.0+   8.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [291]  19.0+  99.0  180.0+ 179.0+ 180.0+ 180.0+  11.0+   0.5   18.0    7.0+
#> [301] 180.0+ 152.0+ 180.0+  93.0  180.0+ 101.0    4.0    5.0    7.0+ 180.0+
#> [311] 180.0+ 180.0+   1.0  180.0+ 180.0+ 171.0  180.0+ 180.0+ 180.0+ 180.0+
#> [321] 180.0+  14.0+ 180.0+   7.0+   2.0   45.0    5.0+ 103.0  180.0+   5.0+
#> [331] 180.0+   8.0+ 180.0+   2.0+  18.0  172.0+ 180.0+ 180.0+   7.0  180.0+
#> [341]   7.0  123.0  180.0+ 180.0+  51.0  180.0+   1.0   60.0   76.0  180.0+
#> [351] 180.0+ 162.0  124.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+  76.0  173.0+
#> [361] 180.0+ 180.0+ 180.0+  28.0  180.0+  16.0+  16.0+ 180.0+ 180.0+   7.0+
#> [371]   3.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0    3.0+  20.0   20.0    8.0 
#> [381]  12.0    4.0+  58.0  180.0+ 180.0+ 180.0+   3.0  175.0  180.0+  10.0+
#> [391] 180.0+ 179.0+   1.0  180.0+  15.0  180.0+  10.0    1.0  180.0+  13.0 
#> [401] 180.0+   4.0+  10.0  104.0+   1.0   57.0  180.0+  11.0    3.0+ 180.0+
#> [411] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 177.0+ 180.0+   5.0   11.0+
#> [421]   4.0+ 180.0+   3.0  180.0+ 180.0+ 180.0+   6.0  180.0+  85.0   17.0+
#> [431] 180.0+   4.0    7.0  180.0+  12.0  180.0+ 180.0+ 180.0+   1.0  180.0+
#> [441] 180.0+ 180.0+ 180.0+  33.0  180.0+ 180.0+ 180.0+   1.0  180.0+   7.0+
#> [451]  79.0    3.0  180.0+ 180.0+ 176.0+  47.0  180.0+  11.0  180.0+   7.0 
#> [461] 180.0+   8.0+   7.0  180.0+  10.0  180.0+ 172.0  119.0   12.0  180.0+
#> [471] 180.0+   1.0   80.0  180.0+   4.0+   2.0  152.0+   3.0   29.0   24.0 
#> [481]  32.0    6.0  180.0+   3.0+ 180.0+ 180.0+ 180.0+ 180.0+  11.0    4.0 
#> [491]  10.0+   6.0    2.0+ 180.0+   1.0  171.0    1.0   43.0    3.0  180.0+
#> [501]   6.0  138.0  180.0+ 180.0+  71.0    8.0   40.0   59.0  161.0   10.0+
#> [511]  93.0  164.0  118.0  173.0  180.0+  37.0  180.0+   7.0+   5.0+ 180.0+
#> [521]   3.0  171.0+ 166.0+  71.0    1.0    3.0+  85.0   10.0    6.0  180.0+
#> [531]   1.0  180.0+ 180.0+ 180.0+ 125.0    4.0    6.0    9.0+ 180.0+ 180.0+
#> [541] 180.0+ 103.0  180.0+   4.0  180.0+ 180.0+ 180.0+  20.0   16.0  180.0+
#> [551] 180.0+ 180.0+ 177.0+   0.5  180.0+ 180.0+   2.0    3.0+  62.0  180.0+
#> [561]   4.0   38.0  180.0+  90.0  180.0+ 180.0+  89.0    4.0   71.0    1.0 
#> [571]  19.0  180.0+ 114.0  180.0+ 180.0+   4.0+ 180.0+   1.0   12.0   16.0+
#> [581]   4.0+ 180.0+   3.0   83.0   88.0  126.0  180.0+ 165.0    3.0  180.0+
#> [591] 180.0+ 180.0+   3.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+  70.0 
#> [601]  43.0  180.0+ 180.0+  14.0   13.0  180.0+ 180.0+  92.0  180.0+   4.0 
#> [611]   3.0+   6.0+  62.0    6.0+   4.0+  65.0   11.0  180.0+  46.0    3.0+
#> [621]   8.0+   4.0  119.0  180.0+   1.0+ 180.0+   8.0  110.0  180.0+ 180.0+
#> [631]   1.0+ 180.0+   8.0   25.0   24.0   50.0  126.0  180.0+ 180.0+   4.0 
#> [641] 178.0+  36.0   89.0   75.0    1.0   33.0  158.0   74.0  180.0+ 168.0 
#> [651]  52.0    7.0    4.0  180.0+   4.0   50.0    1.0+   2.0  179.0+   8.0+
#> [661]  76.0   16.0   67.0    8.0   26.0   53.0    7.0+  69.0  180.0+   3.0 
#> 
#> $weights
#> NULL
#> 
#> $offset
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8321818