Cross-Validated GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::cv.glmnet() from package glmnet.

Details

This learner returns two prediction types:

lp: a vector of linear predictors (relative risk scores), one per observation. Calculated using glmnet::predict.cv.glmnet().

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.cv.glmnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used.

Custom mlr3 parameters

family is set to "cox" and cannot be changed.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.cv_glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

foldid	untyped	NULL		-

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nfolds	integer	10		\([3, \infty)\)

nlambda	integer	100		\([1, \infty)\)

offset	untyped	NULL		-

newoffset	untyped	-		-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	lambda.1se		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

standardize.response	logical	FALSE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.gaussian	character	-	covariance, naive	-

type.logistic	character	Newton	Newton, modified.Newton	-

type.measure	character	deviance	deviance, C	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvCVGlmnet

Methods

Public methods

LearnerSurvCVGlmnet$new()

LearnerSurvCVGlmnet$selected_features()

LearnerSurvCVGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvCVGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvCVGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvCVGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.cv_glmnet")
print(learner)
#> <LearnerSurvCVGlmnet:surv.cv_glmnet>: Regularized Generalized Linear Model
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3proba, mlr3extralearners, glmnet
#> * Predict Types:  [crank], distr, lp
#> * Feature Types: logical, integer, numeric
#> * Properties: selected_features, weights

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#> Measure: Partial Likelihood Deviance 
#> 
#>      Lambda Index Measure     SE Nonzero
#> min 0.00370    43   8.628 0.2123       6
#> 1se 0.04558    16   8.811 0.2163       6
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  32   5      1          0        1   121
#>   [2,]  33   2      0          2        0   150
#>   [3,]  35  10      1          9        0   106
#>   [4,]  34   5      0          5        0   120
#>   [5,]  35   2      0        180        0   121
#>   [6,]  35   2      1          1        1   112
#>   [7,]  37   9      0        180        1   151
#>   [8,]  38  13      1          0        1   161
#>   [9,]  38   2      0        115        0   150
#>  [10,]  36   1      0        180        1   155
#>  [11,]  36   5      1          0        1   115
#>  [12,]  33   6      1          1        1   115
#>  [13,]  38  16      1         10        0   160
#>  [14,]  40  12      1          9        0   153
#>  [15,]  42   3      1          1        1   130
#>  [16,]  37   1      1          0        1   146
#>  [17,]  40   2      1          1        1   148
#>  [18,]  42   2      0        180        1   100
#>  [19,]  38   5      1          3        0   125
#>  [20,]  42   2      0          2        0   140
#>  [21,]  40   6      0        180        1   138
#>  [22,]  42   2      0        180        0   100
#>  [23,]  43   4      1          0        1   130
#>  [24,]  42   4      0        180        0   162
#>  [25,]  40   3      1          1        0   170
#>  [26,]  42  12      1         10        1   170
#>  [27,]  43   2      1          1        1   116
#>  [28,]  42   2      0        180        1   124
#>  [29,]  44   5      1          1        0   170
#>  [30,]  45   3      0        180        1   154
#>  [31,]  41  10      1          8        0   150
#>  [32,]  41  13      1          1        0   140
#>  [33,]  45   9      1          7        0   110
#>  [34,]  45   6      0        180        1   170
#>  [35,]  41   5      1          4        1   141
#>  [36,]  45   2      0        180        1   140
#>  [37,]  46  15      0        180        0   120
#>  [38,]  47   4      1          3        0   118
#>  [39,]  48  15      0        180        1   160
#>  [40,]  45   3      0        150        0   130
#>  [41,]  44   3      1          0        1   180
#>  [42,]  45   4      1          0        0   124
#>  [43,]  47   6      1          0        1   116
#>  [44,]  44   0      1          0        1    96
#>  [45,]  47   4      1          3        1   160
#>  [46,]  43   3      1          0        1   124
#>  [47,]  49   5      0         73        1   136
#>  [48,]  45   5      0          5        0   141
#>  [49,]  46   2      1          1        1   122
#>  [50,]  46   6      1          0        1   100
#>  [51,]  45   5      0        180        1   190
#>  [52,]  46   4      0        180        1   121
#>  [53,]  46  15      0        180        1   120
#>  [54,]  48   3      0        180        0   154
#>  [55,]  47   5      1          3        1   130
#>  [56,]  47   9      1          6        0   170
#>  [57,]  46   3      1          0        1   119
#>  [58,]  47  10      0         10        1   140
#>  [59,]  50   1      1          0        1   129
#>  [60,]  48   2      1          0        0   184
#>  [61,]  47   7      0        180        0   145
#>  [62,]  50   6      1          2        1   140
#>  [63,]  49   7      1          7        1   110
#>  [64,]  46   3      1          1        1   140
#>  [65,]  46   9      1          9        1   122
#>  [66,]  49   2      0          2        0   105
#>  [67,]  51   1      0          1        1   145
#>  [68,]  49  15      1         11        1   160
#>  [69,]  49  23      0        179        1   112
#>  [70,]  52   2      0        180        1   170
#>  [71,]  50   7      1          0        1    92
#>  [72,]  50   4      0          4        1   100
#>  [73,]  51   3      1          2        0   113
#>  [74,]  50   1      1          0        0   150
#>  [75,]  47   8      0        180        0   160
#>  [76,]  51   8      0        180        1   140
#>  [77,]  46   3      0        180        1   120
#>  [78,]  50   4      1          1        0   150
#>  [79,]  48   7      1          0        1   110
#>  [80,]  53   8      0         36        1   160
#>  [81,]  48  17      1         10        0   111
#>  [82,]  47   2      1          1        0   110
#>  [83,]  52   4      1          4        0   152
#>  [84,]  49   9      1          3        0   102
#>  [85,]  53   5      0        180        1   140
#>  [86,]  54  17      1         12        1   102
#>  [87,]  53   7      1          0        0   199
#>  [88,]  51   3      1          1        0   140
#>  [89,]  50   2      0          5        1   106
#>  [90,]  50  10      1          6        0   122
#>  [91,]  53   8      1          7        0   160
#>  [92,]  53   4      0          4        0   140
#>  [93,]  48   6      0        180        0   160
#>  [94,]  48  11      1         10        0   120
#>  [95,]  51  13      0         99        1   160
#>  [96,]  54   9      1          0        1   138
#>  [97,]  55   3      1          1        0   150
#>  [98,]  54  23      1         10        0   131
#>  [99,]  55   6      1          2        1   114
#> [100,]  54   9      1          1        0   130
#> [101,]  52   4      0        180        1   180
#> [102,]  51  13      1         11        0   145
#> [103,]  52   4      0        180        0   183
#> [104,]  50   3      0        174        1   153
#> [105,]  55  28      1         13        1   160
#> [106,]  49   6      1          0        1   130
#> [107,]  50   7      1          1        0   156
#> [108,]  53   8      1          0        1   130
#> [109,]  50   7      1          0        1   127
#> [110,]  52   5      0        175        1   117
#> [111,]  55   2      0          2        0   145
#> [112,]  54   1      0        180        0   162
#> [113,]  54   7      1          0        1   100
#> [114,]  56   2      0        180        0   132
#> [115,]  53  18      1          9        1   150
#> [116,]  55   6      0        180        0   170
#> [117,]  52  16      0         16        0   152
#> [118,]  53  10      1          9        0   172
#> [119,]  52  16      1         14        0   170
#> [120,]  53  15      0         15        1    90
#> [121,]  53   4      0        180        1   150
#> [122,]  54  12      1          0        1   190
#> [123,]  55   2      0        134        1   140
#> [124,]  54   3      0        180        0   128
#> [125,]  56   3      0          8        1   139
#> [126,]  54   2      1          1        0   135
#> [127,]  57   5      1          3        1   138
#> [128,]  56   4      1          0        1   140
#> [129,]  55  11      1          7        0   104
#> [130,]  52  15      1         14        0   130
#> [131,]  57  10      0        180        1   170
#> [132,]  54   5      0        180        1   108
#> [133,]  55   3      1          1        1   156
#> [134,]  57   0      0          0        1   150
#> [135,]  53  21      1         13        1   130
#> [136,]  59   3      1          1        0   172
#> [137,]  58   6      1          0        1    90
#> [138,]  53  15      1         10        1   130
#> [139,]  54  17      1          8        1   227
#> [140,]  55   9      1          2        1   147
#> [141,]  55  13      0        166        1   140
#> [142,]  56   5      0          5        1   150
#> [143,]  57   4      1          2        1   185
#> [144,]  53   7      1          0        1   120
#> [145,]  55   3      1          2        0   140
#> [146,]  55   5      0          5        1   131
#> [147,]  54   7      1          0        1   141
#> [148,]  56   4      0          4        0   164
#> [149,]  59  15      1         10        0   140
#> [150,]  55   5      1          0        0   140
#> [151,]  56   7      1          5        1   120
#> [152,]  55   2      0          2        0   106
#> [153,]  57   1      0        180        0   148
#> [154,]  58   4      1          0        1   160
#> [155,]  57   2      0          2        1   120
#> [156,]  60   5      1          1        0   138
#> [157,]  57   5      0        180        1   130
#> [158,]  55   5      1          0        1   160
#> [159,]  59   6      1          0        1   140
#> [160,]  59   4      1          0        1   152
#> [161,]  61   9      0          9        1   160
#> [162,]  58   4      1          3        0   120
#> [163,]  59   2      1          1        0   140
#> [164,]  58  14      1          6        0   190
#> [165,]  61   4      1          3        0   151
#> [166,]  61   9      1          8        0   150
#> [167,]  61   3      1          2        1   102
#> [168,]  58   1      0          1        1   100
#> [169,]  61  20      1         13        0   130
#> [170,]  57   2      1          0        1   116
#> [171,]  58  10      0         10        1   150
#> [172,]  61   3      0         17        0   143
#> [173,]  57   3      1          2        0   120
#> [174,]  58  19      1         13        1   140
#> [175,]  59   9      1          0        1    80
#> [176,]  58  11      0        172        1   135
#> [177,]  61   4      1          0        1   115
#> [178,]  56   8      1          8        0   120
#> [179,]  61  13      1         12        1   130
#> [180,]  57  15      1         13        1   110
#> [181,]  59   5      1          2        0   182
#> [182,]  58   5      1          1        1   135
#> [183,]  59  10      0        180        0   160
#> [184,]  61   8      0         77        0   120
#> [185,]  62  10      1          0        1   153
#> [186,]  61  28      1          7        0   133
#> [187,]  58   8      1          3        1   150
#> [188,]  61   7      0          7        1   150
#> [189,]  60   7      0          7        0   147
#> [190,]  61   6      0          6        0   134
#> [191,]  59  13      1          2        0   198
#> [192,]  62   4      1          0        0   160
#> [193,]  60  17      1          8        1   140
#> [194,]  58   3      1          0        1   146
#> [195,]  62   4      1          3        0   173
#> [196,]  58   2      0         30        0   202
#> [197,]  59   1      0        180        0   155
#> [198,]  63   6      0         28        1   120
#> [199,]  61  13      0         13        0   120
#> [200,]  57  18      1          9        1    93
#> [201,]  58  11      1          9        0   179
#> [202,]  57   2      1          1        0   159
#> [203,]  62  17      1         10        1   180
#> [204,]  62   1      1          0        1   172
#> [205,]  58   7      0        180        1   150
#> [206,]  63   3      1          1        0   180
#> [207,]  61   7      0        180        0   135
#> [208,]  63   4      1          3        0   222
#> [209,]  63  15      1         10        1   126
#> [210,]  64   4      0        180        0   130
#> [211,]  63   4      1          1        0   155
#> [212,]  59   8      0        180        1   140
#> [213,]  58   9      1          9        0   110
#> [214,]  59   1      0         22        1   162
#> [215,]  58   2      0        180        0   127
#> [216,]  60   7      1          5        1   141
#> [217,]  60   7      0          7        0   140
#> [218,]  59   5      1          1        0   148
#> [219,]  60   7      1          1        1    90
#> [220,]  65  13      0        180        1   100
#> [221,]  63   1      0          1        0   162
#> [222,]  63   1      0          1        0   130
#> [223,]  62   6      0        180        0   170
#> [224,]  61  15      1         13        0   170
#> [225,]  59   4      0          4        0   149
#> [226,]  60   3      0          3        0   168
#> [227,]  64  10      1          9        0   160
#> [228,]  62   6      0          6        0   120
#> [229,]  63  12      1         10        0   200
#> [230,]  60   8      0         17        1   130
#> [231,]  61   6      1          1        1   117
#> [232,]  64  12      1         11        0   160
#> [233,]  64   6      1          0        1   140
#> [234,]  63  10      1          0        1   148
#> [235,]  63  14      1          9        0   123
#> [236,]  65  36      1         11        0   140
#> [237,]  66   3      1          1        0   127
#> [238,]  61  10      1          2        1   194
#> [239,]  63  12      1          9        0   114
#> [240,]  63   7      0        180        0   120
#> [241,]  66   5      1          0        1   110
#> [242,]  65   8      1          0        0   168
#> [243,]  65  10      1          8        1   120
#> [244,]  64   0      0          0        1    90
#> [245,]  60   6      0        180        0   130
#> [246,]  64  21      1         10        0   190
#> [247,]  64   9      0        180        0   150
#> [248,]  61   4      0        180        1   113
#> [249,]  63  16      1          7        1   110
#> [250,]  64   7      0        180        1   120
#> [251,]  66   6      1          1        1   130
#> [252,]  63  12      0         12        1   150
#> [253,]  65   6      0          9        0   112
#> [254,]  65   3      1          0        1    80
#> [255,]  63   2      1          1        0   180
#> [256,]  62  13      1         11        0   180
#> [257,]  67  11      0         11        1   100
#> [258,]  64   2      0          2        0   201
#> [259,]  66  16      1         11        1   169
#> [260,]  61  15      1         10        0   130
#> [261,]  63   9      1          8        1   160
#> [262,]  63   3      1          2        0   120
#> [263,]  63   2      1          0        0   140
#> [264,]  64  19      1          8        1   160
#> [265,]  67   6      0        180        1   170
#> [266,]  68   5      1          4        1   150
#> [267,]  66   7      1          0        1   115
#> [268,]  66  13      1          0        0   118
#> [269,]  64  14      1         13        1   150
#> [270,]  65   3      0          3        0   105
#> [271,]  64   0      0          0        1   148
#> [272,]  67   4      1          3        0   130
#> [273,]  66   6      1          0        1   140
#> [274,]  65   2      1          1        1   170
#> [275,]  67   8      1          1        1   130
#> [276,]  68   5      0          5        1    90
#> [277,]  63  10      0         16        1   160
#> [278,]  66  14      0        180        0   130
#> [279,]  64   1      0          1        1   120
#> [280,]  68  18      0        180        1   260
#> [281,]  63   8      1          1        1   162
#> [282,]  63   1      1          0        1   155
#> [283,]  63  10      0         18        1   130
#> [284,]  68  11      0        180        0   160
#> [285,]  65  15      1         12        1   150
#> [286,]  66  11      1          0        0   100
#> [287,]  65   4      1          2        1   145
#> [288,]  69  12      0         15        1   140
#> [289,]  63   2      0        180        0   150
#> [290,]  69  21      1         10        0   180
#> [291,]  69   6      0        180        1   100
#> [292,]  66   9      1          8        0   130
#> [293,]  63   8      0        180        1   120
#> [294,]  68  14      1         13        1   140
#> [295,]  65   8      1          0        1    90
#> [296,]  66   3      0          3        1   138
#> [297,]  69   1      1          0        0   170
#> [298,]  67   1      0        180        1   160
#> [299,]  68  10      1         10        1   150
#> [300,]  65   1      1          0        0   133
#> [301,]  67   7      1          4        1   130
#> [302,]  63   2      1          0        0    99
#> [303,]  67   2      0        180        0   184
#> [304,]  65  10      1          1        1   148
#> [305,]  69   6      0         99        1   140
#> [306,]  65   4      1          1        0   130
#> [307,]  66   4      0        180        1   130
#> [308,]  70  15      1         12        1   132
#> [309,]  64  11      0         11        0   125
#> [310,]  64   4      0        180        1   140
#> [311,]  64   0      1          0        1   118
#> [312,]  67   2      0         18        0   131
#> [313,]  66   7      1          5        1   131
#> [314,]  65  13      1         12        1   130
#> [315,]  69  17      1         10        0   140
#> [316,]  69   8      0         93        0   140
#> [317,]  66   6      0        180        0   140
#> [318,]  68  18      1          0        1   160
#> [319,]  68   4      0          4        1   190
#> [320,]  71   3      0          5        0   112
#> [321,]  70   7      1          0        1   190
#> [322,]  71  20      1          0        1   160
#> [323,]  67   2      0        180        0   128
#> [324,]  66   1      1          1        1   165
#> [325,]  70   4      1          0        1   180
#> [326,]  70  14      0        171        0   166
#> [327,]  66   4      0        180        0   130
#> [328,]  67   6      1          4        0   130
#> [329,]  68   7      1          0        1   150
#> [330,]  69   3      1          2        0   151
#> [331,]  67  14      1         13        0   130
#> [332,]  65  14      1         13        1   150
#> [333,]  69   8      0        180        1   180
#> [334,]  71   7      0          7        0   230
#> [335,]  66   2      0          2        1   228
#> [336,]  71   6      0         45        1   158
#> [337,]  69   5      0          5        1   142
#> [338,]  71   3      0        103        0   133
#> [339,]  69   3      0          3        1   130
#> [340,]  70  22      1         13        0   103
#> [341,]  69   8      1          5        1   195
#> [342,]  69   6      1          4        1   174
#> [343,]  72   3      1          0        1   132
#> [344,]  72   7      0          7        1   110
#> [345,]  69   8      1          7        1   108
#> [346,]  69  19      0        180        0   130
#> [347,]  69  11      1          0        1   120
#> [348,]  69   4      1          3        0   132
#> [349,]  69   8      1          2        0   121
#> [350,]  67  13      1          9        0   130
#> [351,]  70   9      0        180        1   142
#> [352,]  72   5      1          4        0   170
#> [353,]  67  22      1          1        1   140
#> [354,]  67   4      0         60        1   136
#> [355,]  69   5      0         76        0   120
#> [356,]  72  13      1         11        1   195
#> [357,]  72  30      1          0        1   145
#> [358,]  70   7      0          7        0   102
#> [359,]  68   7      1          2        0   135
#> [360,]  73  20      1          0        1   170
#> [361,]  71   6      0          9        0   120
#> [362,]  70  11      0        180        1   210
#> [363,]  67   8      0        180        1   170
#> [364,]  67   5      1          0        1   147
#> [365,]  67   9      0        180        0   158
#> [366,]  73  13      0        152        1   130
#> [367,]  70   5      0        180        0   150
#> [368,]  67   4      1          1        0   134
#> [369,]  71   1      0        173        1   188
#> [370,]  68  23      0        180        1   220
#> [371,]  72   5      0         28        0   120
#> [372,]  71   5      0        180        0   191
#> [373,]  73   6      0        180        1   117
#> [374,]  69   8      1          1        0   164
#> [375,]  70   4      0        180        0   180
#> [376,]  69   1      1          0        0   155
#> [377,]  73   6      1          0        1   270
#> [378,]  72   8      1          1        1   150
#> [379,]  71   2      1          0        1   180
#> [380,]  73   7      0          7        1   140
#> [381,]  70   3      0          3        1   159
#> [382,]  70  13      1          9        0   100
#> [383,]  73   0      0        180        1   161
#> [384,]  74   8      1          0        1    85
#> [385,]  73   4      0        180        1   154
#> [386,]  69   2      1          0        1   110
#> [387,]  71   3      1          1        0   150
#> [388,]  71  15      1         11        0   165
#> [389,]  74  20      0         20        1   180
#> [390,]  71  20      1         10        0   140
#> [391,]  74   0      1          0        1    90
#> [392,]  73   3      1          0        1   136
#> [393,]  70   5      1          0        1   190
#> [394,]  71  17      1         11        0   160
#> [395,]  71   8      1          7        0   149
#> [396,]  73  10      1          8        0   106
#> [397,]  69  12      1          1        1   149
#> [398,]  70  26      1         11        1   120
#> [399,]  74   4      0          4        0   120
#> [400,]  72   5      1          3        1   160
#> [401,]  70   3      0        180        1   154
#> [402,]  73   6      0        180        0   110
#> [403,]  72  15      1          0        1   150
#> [404,]  71   7      1          2        0   143
#> [405,]  72   8      1          0        1   140
#> [406,]  73  17      1         11        0   140
#> [407,]  71  13      1          8        0   121
#> [408,]  69   2      1          1        1    80
#> [409,]  70   4      1          0        1   140
#> [410,]  72  10      1          8        1   153
#> [411,]  69   7      0        180        1   144
#> [412,]  72  15      1         13        0   156
#> [413,]  71  10      1          9        1   120
#> [414,]  73  10      1          9        1   146
#> [415,]  73  10      1         10        1   120
#> [416,]  75   9      1          7        0   140
#> [417,]  75  13      1          1        1   130
#> [418,]  71  11      1          8        0   110
#> [419,]  72  15      1         12        1   120
#> [420,]  73  10      1          8        0   120
#> [421,]  70   7      1          4        0   184
#> [422,]  72   1      1          1        0   168
#> [423,]  72   7      0         57        1   145
#> [424,]  73  10      0        180        0   162
#> [425,]  70   3      0          3        0   150
#> [426,]  73   5      1          3        1   112
#> [427,]  76  25      1         12        1   170
#> [428,]  73  12      1         12        1   140
#> [429,]  72   2      0        180        0   120
#> [430,]  75   1      0        180        1   140
#> [431,]  72   4      1          0        1   197
#> [432,]  71   3      1          0        0   144
#> [433,]  73   5      0        180        0   126
#> [434,]  73   4      0        180        0   124
#> [435,]  76   3      1          0        1   120
#> [436,]  72   5      0        180        0   154
#> [437,]  77  11      0         11        1   150
#> [438,]  72   7      1          2        0   142
#> [439,]  73  15      0         15        1   160
#> [440,]  71  16      0        180        0   140
#> [441,]  73  10      1         10        0   124
#> [442,]  74   2      1          1        0   140
#> [443,]  74  19      1          4        1   200
#> [444,]  73   6      0          6        1   114
#> [445,]  75  23      1         14        1   110
#> [446,]  74   2      0        180        0   190
#> [447,]  72   4      1          3        0   160
#> [448,]  76  17      1          0        1   200
#> [449,]  75   4      1          0        1   122
#> [450,]  75   7      0          7        0   190
#> [451,]  73  13      1         11        0   195
#> [452,]  74   8      1          0        1   105
#> [453,]  76  13      1          8        1   148
#> [454,]  75   4      1          2        1   188
#> [455,]  74   6      0        180        0   160
#> [456,]  76   4      0          4        1   155
#> [457,]  75   1      0          1        1   125
#> [458,]  73   1      0         52        1   105
#> [459,]  73   0      0        180        0   156
#> [460,]  72   5      0        180        0   120
#> [461,]  76  44      1         10        0   105
#> [462,]  76   5      0        180        0   185
#> [463,]  74  10      1          0        1   135
#> [464,]  76   5      1          0        1   167
#> [465,]  74   8      1          8        1   170
#> [466,]  75   9      0        180        1   140
#> [467,]  77   5      1          0        0   123
#> [468,]  77  12      1          9        1   100
#> [469,]  73  10      1          9        0   146
#> [470,]  77  12      0        180        0   130
#> [471,]  77   1      1          0        1    90
#> [472,]  76  12      1         11        1   120
#> [473,]  75   3      1          1        1   171
#> [474,]  74   2      1          0        1   130
#> [475,]  78  18      0         18        1   144
#> [476,]  73  11      1          2        1   110
#> [477,]  78   7      0          7        1   133
#> [478,]  76  13      1          1        1   170
#> [479,]  78  32      1          9        1   198
#> [480,]  79   6      0        180        0   170
#> [481,]  80  10      1          6        1   147
#> [482,]  78   0      0        180        1   212
#> [483,]  75   5      0        119        1   150
#> [484,]  75  12      1          1        1   120
#> [485,]  78  15      0        180        1   270
#> [486,]  74  10      1          8        0   135
#> [487,]  76   1      0          1        1    83
#> [488,]  79   4      0         80        0   145
#> [489,]  78  12      1          9        0   150
#> [490,]  75   4      1          0        0   212
#> [491,]  77   2      1          0        1   143
#> [492,]  78  10      0        180        1   130
#> [493,]  76  11      1          0        0   120
#> [494,]  75  11      1          4        0   162
#> [495,]  75   3      0          3        0     0
#> [496,]  80   9      0         23        1   128
#> [497,]  80   6      0          6        1   150
#> [498,]  78   6      1          0        1   240
#> [499,]  76   3      1          0        1   140
#> [500,]  78  11      1          1        1   140
#> [501,]  79  11      0        180        0   160
#> [502,]  79   2      1          0        1   121
#> [503,]  78  14      1          0        1   140
#> [504,]  81   1      0          1        0   130
#> [505,]  78  11      1          8        1   118
#> [506,]  79   4      0          4        1   125
#> [507,]  76  10      1          8        0   180
#> [508,]  77   6      0          6        1   107
#> [509,]  78  11      0        180        1   135
#> [510,]  77  31      1          3        1   161
#> [511,]  76   1      0          1        1    90
#> [512,]  79   3      0          3        0   120
#> [513,]  77   7      0        180        1   170
#> [514,]  77   6      0          6        1   144
#> [515,]  79   4      1          0        1   120
#> [516,]  81   1      0        180        0   120
#> [517,]  80  15      1         12        1   150
#> [518,]  77   9      1          4        0   141
#> [519,]  82   5      0          8        1   120
#> [520,]  78   4      0         59        1   112
#> [521,]  79  10      0         10        1   120
#> [522,]  80  15      1          0        1    90
#> [523,]  81   4      1          2        1   126
#> [524,]  79  28      0        164        0   100
#> [525,]  80   9      0        118        1   186
#> [526,]  80   6      0        173        1   160
#> [527,]  79   1      0         37        1   140
#> [528,]  81   3      0        180        0   184
#> [529,]  78   7      0          7        1   147
#> [530,]  78  15      0         15        0   165
#> [531,]  79   3      0          3        1   101
#> [532,]  78  26      1          5        0   194
#> [533,]  81   4      1          1        1   104
#> [534,]  78  20      1          0        1   109
#> [535,]  80   1      0          1        0   100
#> [536,]  78   3      1          1        1   152
#> [537,]  77  10      1          8        1   130
#> [538,]  77   5      0         85        0   188
#> [539,]  80   2      1          1        0   168
#> [540,]  79   6      0          6        0   152
#> [541,]  80   6      1          0        1   119
#> [542,]  78   2      0        180        0   148
#> [543,]  80   5      0          5        1   130
#> [544,]  82   1      1          0        1    82
#> [545,]  77   4      0        180        1    98
#> [546,]  81   1      0        108        0   129
#> [547,]  79   1      0        125        0   193
#> [548,]  79   4      0          4        1   121
#> [549,]  83   9      1          5        1   170
#> [550,]  82   5      0        180        0   110
#> [551,]  83   5      0        180        0   148
#> [552,]  79   7      1          6        0   130
#> [553,]  81   5      0        177        0    41
#> [554,]  80  11      1          8        0   170
#> [555,]  79   4      0          4        1   183
#> [556,]  78   9      1          4        1   120
#> [557,]  82   8      1          1        0   128
#> [558,]  81  15      0        180        1   140
#> [559,]  84   5      1          1        1    85
#> [560,]  81  20      1          9        0   170
#> [561,]  81  16      0         16        1   110
#> [562,]  80  11      1          8        0   110
#> [563,]  81   8      0        180        0   146
#> [564,]  80   8      1          7        0   160
#> [565,]  79   7      0        177        0   197
#> [566,]  79   0      1          0        1    96
#> [567,]  85   4      0        180        0    90
#> [568,]  83   2      0          2        1   155
#> [569,]  82   6      0        128        1   100
#> [570,]  80   3      1          1        1   230
#> [571,]  84   4      0          4        1    85
#> [572,]  81   1      0          1        1   150
#> [573,]  84   1      0         38        1   205
#> [574,]  83   3      0        180        0   174
#> [575,]  81   4      0         90        1   138
#> [576,]  79   9      1          8        0   150
#> [577,]  85   3      1          2        1   160
#> [578,]  84   4      0         89        1   129
#> [579,]  80   2      1          0        1   130
#> [580,]  79   4      0          4        1    60
#> [581,]  80   6      0         71        1   189
#> [582,]  83   9      0        180        0   198
#> [583,]  79  14      1          0        0   110
#> [584,]  83   3      0        114        0    98
#> [585,]  81  14      1         12        1   128
#> [586,]  83   2      0        154        0   130
#> [587,]  83   1      0        180        0   160
#> [588,]  81   4      0          4        0   175
#> [589,]  84  15      1         13        1   110
#> [590,]  82   5      1          0        1   146
#> [591,]  81   4      0          4        0   160
#> [592,]  83  12      1          2        1   170
#> [593,]  82  15      1          0        0   183
#> [594,]  80   2      0         88        0   135
#> [595,]  83   7      0        126        0   135
#> [596,]  81  16      1          9        0   180
#> [597,]  86   3      0          3        1   140
#> [598,]  84   3      0        180        1   120
#> [599,]  81  13      0        180        0   152
#> [600,]  85   3      0          3        1   118
#> [601,]  81   2      1          0        1   118
#> [602,]  81   4      0        180        0   160
#> [603,]  82   1      0        180        1   193
#> [604,]  87   2      0          5        1   137
#> [605,]  86  12      1          0        1   132
#> [606,]  82  14      1         11        1   103
#> [607,]  84  16      0         70        1   150
#> [608,]  83  19      0         43        0   150
#> [609,]  84   3      1          2        0   125
#> [610,]  83  10      1          0        1   190
#> [611,]  88  14      1          3        1   130
#> [612,]  87   2      0        180        0   113
#> [613,]  84   3      0        180        1   170
#> [614,]  82   4      0          4        0   130
#> [615,]  85   3      0          3        1   113
#> [616,]  86   6      0          6        1   117
#> [617,]  86   6      1          1        0   112
#> [618,]  88   4      0          4        0   100
#> [619,]  83  20      1          3        1   150
#> [620,]  88   4      0          4        1   115
#> [621,]  85  22      0         22        1   184
#> [622,]  86   9      1          7        1   142
#> [623,]  86   6      0         46        0   173
#> [624,]  88   3      0        115        0   110
#> [625,]  88   4      0          4        0    86
#> [626,]  89   4      0          4        1   153
#> [627,]  87   6      0        180        1   110
#> [628,]  87   1      0          1        0   170
#> [629,]  84   8      0        180        1   119
#> [630,]  85   8      0          8        1   136
#> [631,]  87  29      0         29        1    97
#> [632,]  89  10      0         46        1   170
#> [633,]  90  14      0         14        1   100
#> [634,]  88   1      0          1        0   135
#> [635,]  91   8      0          8        0   100
#> [636,]  87   2      0        180        0   160
#> [637,]  87   6      1          0        0   125
#> [638,]  91  10      0        145        0   135
#> [639,]  86   3      1          0        1    80
#> [640,]  88   7      0         24        0   119
#> [641,]  90  11      1         10        1   186
#> [642,]  86  10      0        180        1   137
#> [643,]  86   9      1          7        0   130
#> [644,]  91   1      0          1        1    74
#> [645,]  87  43      0        178        1   130
#> [646,]  87   5      0         36        1   150
#> [647,]  88   3      1          2        0   159
#> [648,]  89   3      1          1        1   160
#> [649,]  91   3      0         33        1   137
#> [650,]  89  12      1          0        1   130
#> [651,]  91   5      0        169        1   176
#> [652,]  89  52      0         52        1   130
#> [653,]  89   4      0          4        1   159
#> [654,]  91   0      0          0        0     0
#> [655,]  89  14      0        180        1    84
#> [656,]  90  18      0        180        0   188
#> [657,]  90  19      1         11        1   129
#> [658,]  90   1      0          1        1   118
#> [659,]  91   2      0          2        1   116
#> [660,]  92   4      0         76        1   149
#> [661,]  91   1      0        180        0   158
#> [662,]  96   3      0         12        1    97
#> [663,]  95   8      1          5        1   150
#> [664,]  94   3      0         26        1   144
#> [665,]  91  12      0         53        1   212
#> [666,]  93   0      1          0        1   122
#> [667,]  92   5      0         69        0   139
#> [668,]  92   2      0          2        0   112
#> [669,]  93   4      0        180        1   135
#> [670,]  96   3      1          0        1   104
#> 
#> $y
#>   [1]   5.0+   2.0+ 180.0+   5.0+ 180.0+   2.0+ 180.0+ 180.0+ 115.0  180.0+
#>  [11]   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 180.0+   5.0+   2.0+
#>  [21] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 180.0+ 155.0+ 180.0+
#>  [31] 180.0+ 180.0+ 180.0+ 180.0+   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 150.0 
#>  [41] 180.0+ 180.0+   6.0+ 180.0+ 180.0+ 180.0+  73.0    5.0+ 161.0+ 180.0+
#>  [51] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  10.0+ 172.0+ 180.0+
#>  [61] 180.0+ 180.0+   7.0  180.0+ 180.0+   2.0    1.0  179.0+ 179.0+ 180.0+
#>  [71] 180.0+   4.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   7.0   36.0 
#>  [81]  88.0+ 180.0+   4.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0  180.0+
#>  [91] 180.0+   4.0+ 180.0+ 180.0+  99.0  180.0+ 180.0+ 152.0+   6.0+ 180.0+
#> [101] 180.0+  13.0+ 180.0+ 174.0+  28.0    6.0+ 180.0+ 180.0+ 180.0+ 175.0+
#> [111]   2.0  180.0+   7.0+ 180.0+ 180.0+ 180.0+  16.0+ 180.0+  16.0   15.0+
#> [121] 180.0+  12.0+ 134.0+ 180.0+   8.0  180.0+ 140.0  165.0  180.0+ 180.0+
#> [131] 180.0+ 180.0+ 180.0+   0.5  180.0+ 180.0+ 180.0+ 180.0+ 171.0+  15.0 
#> [141] 166.0+   5.0+   4.0+ 180.0+ 180.0+   5.0+ 180.0+   4.0+ 180.0+ 180.0+
#> [151] 180.0+   2.0+ 180.0+ 180.0+   2.0  180.0+ 180.0+ 180.0+  64.0  180.0+
#> [161]   9.0+ 180.0+ 180.0+ 171.0+ 180.0+ 180.0+   3.0    1.0  180.0+ 180.0+
#> [171]  10.0+  17.0    3.0+  19.0    9.0+ 172.0+ 180.0+   8.0  180.0+  15.0 
#> [181] 180.0+ 180.0+ 180.0+  77.0  180.0+  94.0  180.0+   7.0    7.0+   6.0 
#> [191] 180.0+ 180.0+ 180.0+   3.0+ 180.0+  30.0  180.0+  28.0   13.0+  18.0 
#> [201] 180.0+ 180.0+ 180.0+   1.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [211]   4.0+ 180.0+   9.0   22.0  180.0+  84.0    7.0+ 180.0+ 180.0+ 180.0+
#> [221]   1.0    1.0  180.0+ 180.0+   4.0+   3.0+ 167.0    6.0+ 180.0+  17.0 
#> [231] 180.0+  12.0  180.0+ 180.0+  14.0+  36.0    3.0+  88.0   12.0  180.0+
#> [241] 180.0+ 180.0+ 180.0+   0.5  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [251] 180.0+  12.0    9.0    3.0  180.0+ 180.0+  11.0+   2.0+ 180.0+ 180.0+
#> [261] 180.0+   3.0+   2.0+ 103.0  180.0+   5.0+ 179.0+ 166.0+  14.0+   3.0 
#> [271]   0.5+ 180.0+ 180.0+ 175.0+   8.0    5.0   16.0  180.0+   1.0  180.0+
#> [281] 180.0+   1.0+  18.0  180.0+  15.0+ 180.0+   4.0+  15.0  180.0+ 174.0+
#> [291] 180.0+ 180.0+ 180.0+ 180.0+   8.0+   3.0  175.0  180.0+  10.0  180.0+
#> [301] 180.0+ 180.0+ 180.0+ 180.0+  99.0  180.0+ 180.0+ 180.0+  11.0+ 180.0+
#> [311]   0.5   18.0    7.0+ 180.0+ 180.0+  93.0  180.0+  18.0+   4.0    5.0 
#> [321]   7.0+ 180.0+ 180.0+   1.0  180.0+ 171.0  180.0+   6.0  180.0+ 180.0+
#> [331] 180.0+  14.0+ 180.0+   7.0+   2.0   45.0    5.0+ 103.0    3.0+ 180.0+
#> [341] 180.0+  97.0  180.0+   7.0    8.0+ 180.0+ 180.0+ 180.0+   8.0+  13.0+
#> [351] 180.0+ 180.0+  51.0   60.0   76.0  132.0  162.0    7.0+   7.0+ 124.0 
#> [361]   9.0  180.0+ 180.0+ 180.0+ 180.0+ 152.0  180.0+  76.0  173.0+ 180.0+
#> [371]  28.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+   6.0  180.0+ 180.0+   7.0+
#> [381]   3.0+  13.0+ 180.0+ 180.0+ 180.0+   2.0    3.0+ 180.0+  20.0   20.0 
#> [391]   0.5  180.0+ 180.0+ 180.0+   8.0   87.0   12.0  180.0+   4.0+ 180.0+
#> [401] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 175.0    2.0  180.0+  10.0+
#> [411] 180.0+ 180.0+ 179.0+ 180.0+  15.0  180.0+  13.0  180.0+ 180.0+  10.0 
#> [421] 104.0+   1.0   57.0  180.0+   3.0+   5.0  180.0+  12.0  180.0+ 180.0+
#> [431] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  11.0+   7.0   15.0+ 180.0+
#> [441]  10.0  180.0+ 180.0+   6.0  180.0+ 180.0+ 180.0+  17.0+   4.0    7.0 
#> [451] 180.0+ 180.0+ 180.0+  46.0  180.0+   4.0    1.0   52.0  180.0+ 180.0+
#> [461] 180.0+ 180.0+ 180.0+ 180.0+   8.0  180.0+   5.0  180.0+ 180.0+ 180.0+
#> [471]   1.0   12.0    3.0  176.0+  18.0   11.0    7.0  180.0+  32.0  180.0+
#> [481]  10.0  180.0+ 119.0   12.0  180.0+ 180.0+   1.0   80.0  180.0+   4.0+
#> [491]   2.0  180.0+  11.0  152.0+   3.0   23.0    6.0  180.0+   3.0+ 180.0+
#> [501] 180.0+ 180.0+ 180.0+   1.0   11.0    4.0   10.0+   6.0  180.0+ 171.0 
#> [511]   1.0    3.0  180.0+   6.0  138.0  180.0+ 180.0+  71.0    8.0   59.0 
#> [521]  10.0+ 180.0+  93.0  164.0  118.0  173.0   37.0  180.0+   7.0+  15.0+
#> [531]   3.0  171.0+  71.0   20.0+   1.0    3.0+  10.0   85.0   10.0    6.0+
#> [541]   6.0  180.0+   5.0    1.0  180.0+ 108.0  125.0    4.0    9.0+ 180.0+
#> [551] 180.0+ 180.0+ 177.0+ 169.0    4.0  180.0+ 180.0+ 180.0+ 180.0+  20.0 
#> [561]  16.0  180.0+ 180.0+ 180.0+ 177.0+   0.5  180.0+   2.0  128.0    3.0+
#> [571]   4.0    1.0   38.0  180.0+  90.0  180.0+ 180.0+  89.0  180.0+   4.0 
#> [581]  71.0  180.0+ 180.0+ 114.0  180.0+ 154.0  180.0+   4.0+ 180.0+   5.0+
#> [591]   4.0+  77.0   83.0   88.0  126.0  180.0+   3.0  180.0+ 180.0+   3.0+
#> [601] 180.0+ 180.0+ 180.0+   5.0  180.0+ 174.0   70.0   43.0  180.0+ 180.0+
#> [611]  14.0  180.0+ 180.0+   4.0    3.0+   6.0+   6.0+   4.0+  20.0    4.0 
#> [621]  22.0   11.0   46.0  115.0    4.0    4.0  180.0+   1.0+ 180.0+   8.0 
#> [631]  29.0   46.0   14.0    1.0+   8.0  180.0+  25.0  145.0    3.0   24.0 
#> [641]  11.0  180.0+ 180.0+   1.0  178.0+  36.0   75.0    3.0+  33.0  180.0+
#> [651] 169.0   52.0    4.0    0.5  180.0+ 180.0+ 180.0+   1.0+   2.0   76.0 
#> [661] 180.0+  12.0    8.0   26.0   53.0    0.5   69.0    2.0  180.0+   3.0 
#> 
#> $weights
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8465557