Cross-Validated GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::cv.glmnet() from package glmnet.

Details

This learner returns two prediction types:

lp: a vector of linear predictors (relative risk scores), one per observation. Calculated using glmnet::predict.cv.glmnet().

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.cv.glmnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used.

Custom mlr3 parameters

family is set to "cox" and cannot be changed.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.cv_glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

foldid	untyped	NULL		-

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nfolds	integer	10		\([3, \infty)\)

nlambda	integer	100		\([1, \infty)\)

offset	untyped	NULL		-

newoffset	untyped	-		-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	lambda.1se		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

standardize.response	logical	FALSE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.gaussian	character	-	covariance, naive	-

type.logistic	character	Newton	Newton, modified.Newton	-

type.measure	character	deviance	deviance, C	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvCVGlmnet

Methods

Public methods

LearnerSurvCVGlmnet$new()

LearnerSurvCVGlmnet$selected_features()

LearnerSurvCVGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvCVGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvCVGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvCVGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.cv_glmnet")
print(learner)
#> <LearnerSurvCVGlmnet:surv.cv_glmnet>: Regularized Generalized Linear Model
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3proba, mlr3extralearners, glmnet
#> * Predict Types:  [crank], distr, lp
#> * Feature Types: logical, integer, numeric
#> * Properties: selected_features, weights

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#> Measure: Partial Likelihood Deviance 
#> 
#>      Lambda Index Measure     SE Nonzero
#> min 0.00334    44   8.814 0.2337       6
#> 1se 0.04959    15   9.025 0.1956       5
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  28   9      0        180        1   107
#>   [2,]  32   5      1          0        1   121
#>   [3,]  35   5      1          2        0   172
#>   [4,]  34   5      0          5        0   120
#>   [5,]  35   2      0        180        0   121
#>   [6,]  35   2      1          1        1   112
#>   [7,]  38  13      1          0        1   161
#>   [8,]  38   2      0        115        0   150
#>   [9,]  36   1      0        180        1   155
#>  [10,]  35   0      0        180        1   119
#>  [11,]  38  12      1          8        1   120
#>  [12,]  36   5      1          0        1   115
#>  [13,]  33   6      1          1        1   115
#>  [14,]  38  16      1         10        0   160
#>  [15,]  38  12      1         11        1    92
#>  [16,]  40  12      1          9        0   153
#>  [17,]  37   1      1          0        1   146
#>  [18,]  40   2      1          1        1   148
#>  [19,]  42   2      0        180        1   100
#>  [20,]  42   2      0          2        0   140
#>  [21,]  40   6      0        180        1   138
#>  [22,]  40  11      1         10        1   120
#>  [23,]  42   2      0        180        0   100
#>  [24,]  40   1      1          0        1   145
#>  [25,]  43   4      1          0        1   130
#>  [26,]  40   3      1          1        0   170
#>  [27,]  42  12      1         10        1   170
#>  [28,]  43   2      1          1        1   116
#>  [29,]  42   2      0        180        1   124
#>  [30,]  44   5      1          1        0   170
#>  [31,]  45   3      0        180        1   154
#>  [32,]  41  10      1          8        0   150
#>  [33,]  41  13      1          1        0   140
#>  [34,]  45   9      1          7        0   110
#>  [35,]  45   6      0        180        1   170
#>  [36,]  41   5      1          4        1   141
#>  [37,]  44   2      1          1        1   150
#>  [38,]  45   2      0        180        1   140
#>  [39,]  46  15      0        180        0   120
#>  [40,]  46   2      1          1        0   126
#>  [41,]  47   4      1          3        0   118
#>  [42,]  45   3      0        150        0   130
#>  [43,]  44   3      1          0        1   180
#>  [44,]  43  29      0        180        1   180
#>  [45,]  45   4      1          0        0   124
#>  [46,]  43  10      0        180        0   185
#>  [47,]  47   6      1          0        1   116
#>  [48,]  44   0      1          0        1    96
#>  [49,]  47   4      1          3        1   160
#>  [50,]  43   3      1          0        1   124
#>  [51,]  45   8      1          0        1   117
#>  [52,]  49   5      0         73        1   136
#>  [53,]  46   2      1          1        1   122
#>  [54,]  46   6      1          0        1   100
#>  [55,]  44   4      1          0        1   114
#>  [56,]  47   2      0        180        0   108
#>  [57,]  46   5      1          3        0   130
#>  [58,]  46   4      0        180        1   121
#>  [59,]  44   2      0        180        0   142
#>  [60,]  46  15      0        180        1   120
#>  [61,]  45   9      1          0        1   145
#>  [62,]  48   3      0        180        0   154
#>  [63,]  48  12      1         11        0   200
#>  [64,]  47   5      1          3        1   130
#>  [65,]  46   3      1          0        1   119
#>  [66,]  47  10      0         10        1   140
#>  [67,]  48   2      1          0        0   184
#>  [68,]  47   7      0        180        0   145
#>  [69,]  50   6      1          2        1   140
#>  [70,]  46   9      1          9        1   122
#>  [71,]  50   7      0        180        1   110
#>  [72,]  51   1      0          1        1   145
#>  [73,]  49  15      1         11        1   160
#>  [74,]  47   2      0        180        0   150
#>  [75,]  46   6      1          0        1   156
#>  [76,]  52   2      0        180        1   170
#>  [77,]  50   7      1          0        1    92
#>  [78,]  51   3      1          2        0   113
#>  [79,]  50   9      0        180        0   130
#>  [80,]  49   7      1          4        1    90
#>  [81,]  47   6      0        180        1   162
#>  [82,]  51   8      0        180        1   140
#>  [83,]  46   1      1          1        0   145
#>  [84,]  48   7      1          0        1   110
#>  [85,]  53   8      0         36        1   160
#>  [86,]  47   2      1          1        0   110
#>  [87,]  52   4      1          4        0   152
#>  [88,]  53   5      0        180        1   140
#>  [89,]  53   5      0         77        0   159
#>  [90,]  53   7      1          0        0   199
#>  [91,]  54   6      1          3        0   129
#>  [92,]  51   3      1          1        0   140
#>  [93,]  50  10      1          6        0   122
#>  [94,]  53   8      1          7        0   160
#>  [95,]  48   3      1          2        0   150
#>  [96,]  51  25      1          1        0   202
#>  [97,]  49   5      1          2        1   150
#>  [98,]  52  14      1          7        1   200
#>  [99,]  48   6      0        180        0   160
#> [100,]  48  11      1         10        0   120
#> [101,]  53   4      1          0        1   156
#> [102,]  51  13      0         99        1   160
#> [103,]  54   9      1          0        1   138
#> [104,]  49  16      0         16        0   125
#> [105,]  54  23      1         10        0   131
#> [106,]  52   7      1          2        0   154
#> [107,]  55   6      1          2        1   114
#> [108,]  55   4      1          2        0   150
#> [109,]  51  13      1         11        0   145
#> [110,]  55  28      1         13        1   160
#> [111,]  49   6      1          0        1   130
#> [112,]  49   1      0          1        1   110
#> [113,]  53   9      0          9        1    95
#> [114,]  53   8      1          0        1   130
#> [115,]  50   7      1          0        1   127
#> [116,]  56   4      1          1        1   130
#> [117,]  55   2      0          2        0   145
#> [118,]  54   1      0        180        0   162
#> [119,]  54   7      1          0        1   100
#> [120,]  56   3      0        180        1   193
#> [121,]  56   2      0        180        0   132
#> [122,]  55   5      1          4        1   120
#> [123,]  52   8      0        180        0   119
#> [124,]  55   6      0        180        0   170
#> [125,]  52  16      0         16        0   152
#> [126,]  52  16      1         14        0   170
#> [127,]  53  15      0         15        1    90
#> [128,]  53   4      0        180        1   150
#> [129,]  55   6      1          5        1   138
#> [130,]  54  12      1          0        1   190
#> [131,]  56   3      0          8        1   139
#> [132,]  57   3      0          3        0   120
#> [133,]  54   7      1          2        0   129
#> [134,]  54   2      1          1        0   135
#> [135,]  54   2      1          1        1   176
#> [136,]  57   5      1          3        1   138
#> [137,]  57   1      0        180        1   156
#> [138,]  57   1      0          1        1   100
#> [139,]  56   4      1          0        1   140
#> [140,]  52  15      1         14        0   130
#> [141,]  53   3      1          0        1   200
#> [142,]  57  10      0        180        1   170
#> [143,]  54   5      0        180        1   108
#> [144,]  57   4      0        180        1   119
#> [145,]  58   6      1          0        1    90
#> [146,]  54  17      1          8        1   227
#> [147,]  55   9      1          2        1   147
#> [148,]  55  13      0        166        1   140
#> [149,]  56   5      0          5        1   150
#> [150,]  54  23      1          8        0   120
#> [151,]  53   7      1          0        1   120
#> [152,]  55   3      1          2        0   140
#> [153,]  55   5      0          5        1   131
#> [154,]  59  15      1         10        0   140
#> [155,]  58   1      1          1        1   200
#> [156,]  55   5      1          0        0   140
#> [157,]  56   7      1          5        1   120
#> [158,]  59   9      1          1        1   125
#> [159,]  60  11      1          9        0   106
#> [160,]  59   3      0        180        0   120
#> [161,]  58   4      1          0        1   160
#> [162,]  60   5      1          1        0   138
#> [163,]  57   5      0        180        1   130
#> [164,]  58  11      1          9        1   124
#> [165,]  59   6      1          0        1   140
#> [166,]  59   5      0        180        1   155
#> [167,]  59   4      1          0        1   152
#> [168,]  58  26      1          0        1   189
#> [169,]  58   4      1          3        0   120
#> [170,]  60   0      1          0        1    80
#> [171,]  59   2      1          1        0   140
#> [172,]  58   8      0        161        1   140
#> [173,]  58  14      1          6        0   190
#> [174,]  61   3      1          2        1   102
#> [175,]  58   1      0          1        1   100
#> [176,]  57  13      1         10        0   110
#> [177,]  57   2      1          0        1   116
#> [178,]  58  10      0         10        1   150
#> [179,]  61   3      0         17        0   143
#> [180,]  56  14      0         45        0   130
#> [181,]  57   3      1          2        0   120
#> [182,]  58  19      1         13        1   140
#> [183,]  56  13      1          6        1   158
#> [184,]  56  18      1         11        1   165
#> [185,]  59   9      1          0        1    80
#> [186,]  55   4      1          3        1   160
#> [187,]  58  11      0        172        1   135
#> [188,]  60  12      1          0        1   114
#> [189,]  56   8      1          8        0   120
#> [190,]  61  13      1         12        1   130
#> [191,]  59  11      1          8        1   190
#> [192,]  57   1      0          1        0   126
#> [193,]  57  15      1         13        1   110
#> [194,]  59   5      1          2        0   182
#> [195,]  58   5      1          1        1   135
#> [196,]  59  10      0        180        0   160
#> [197,]  61   8      0         77        0   120
#> [198,]  58   8      1          5        0   152
#> [199,]  57   3      1          0        0   100
#> [200,]  61  18      0        170        0   140
#> [201,]  61  28      1          7        0   133
#> [202,]  57   7      0        169        0   180
#> [203,]  60   7      0          7        0   147
#> [204,]  61   6      0          6        0   134
#> [205,]  59  13      1          2        0   198
#> [206,]  57  12      1          9        1   120
#> [207,]  62   4      1          0        0   160
#> [208,]  60  17      1          8        1   140
#> [209,]  58   3      1          0        1   146
#> [210,]  58   2      0         30        0   202
#> [211,]  59  16      1          9        1   133
#> [212,]  61   5      0          5        1   110
#> [213,]  57  18      1          9        1    93
#> [214,]  61   5      0          5        1   160
#> [215,]  58  11      1          9        0   179
#> [216,]  62  17      1         10        1   180
#> [217,]  62   1      1          0        1   172
#> [218,]  58   7      0        180        1   150
#> [219,]  63   3      1          1        0   180
#> [220,]  63   1      0        180        1   130
#> [221,]  63   4      1          3        0   222
#> [222,]  63  15      1         10        1   126
#> [223,]  63   4      1          1        0   155
#> [224,]  59   8      0        180        1   140
#> [225,]  58   9      1          9        0   110
#> [226,]  59   1      0         22        1   162
#> [227,]  58   2      0        180        0   127
#> [228,]  59   4      0        180        0   196
#> [229,]  60   7      1          5        1   141
#> [230,]  60   7      1          1        1    90
#> [231,]  65  13      0        180        1   100
#> [232,]  63   1      0          1        0   162
#> [233,]  63   1      0          1        0   130
#> [234,]  61  15      1         13        0   170
#> [235,]  64  10      1          9        0   160
#> [236,]  59  10      0        180        1   130
#> [237,]  61   6      1          1        1   117
#> [238,]  66   1      1          0        1   120
#> [239,]  64   6      1          0        1   140
#> [240,]  63  14      1          9        0   123
#> [241,]  65  36      1         11        0   140
#> [242,]  63   4      1          3        0   162
#> [243,]  66   3      1          1        0   127
#> [244,]  61  10      1          2        1   194
#> [245,]  64  32      1          9        1   160
#> [246,]  63  12      1          9        0   114
#> [247,]  63   7      0        180        0   120
#> [248,]  65   8      1          0        0   168
#> [249,]  65  10      1          8        1   120
#> [250,]  64   0      0          0        1    90
#> [251,]  64  21      1         10        0   190
#> [252,]  61   4      0        180        1   113
#> [253,]  66   6      1          1        1   130
#> [254,]  63  12      0         12        1   150
#> [255,]  62   3      1          1        1   199
#> [256,]  65   3      1          0        1    80
#> [257,]  63   5      1          4        0   170
#> [258,]  63   2      1          1        0   180
#> [259,]  64   2      0          2        0   201
#> [260,]  66  18      1          5        0   142
#> [261,]  66  16      1         11        1   169
#> [262,]  62   9      0        180        0   145
#> [263,]  61  14      1          5        0   140
#> [264,]  61  15      1         10        0   130
#> [265,]  63   9      1          8        1   160
#> [266,]  63   3      1          2        0   120
#> [267,]  65   8      1          0        1   140
#> [268,]  64   6      1          0        1   125
#> [269,]  66   7      1          0        1   115
#> [270,]  66  13      1          0        0   118
#> [271,]  64  14      1         13        1   150
#> [272,]  64   0      0          0        1   148
#> [273,]  67   4      1          3        0   130
#> [274,]  66   3      1          0        1   135
#> [275,]  68   1      0        180        1   166
#> [276,]  63   7      1          0        0   162
#> [277,]  68   5      0          5        1    90
#> [278,]  64   1      0          1        1   120
#> [279,]  68  18      0        180        1   260
#> [280,]  65  17      1         14        1   100
#> [281,]  63   8      1          1        1   162
#> [282,]  65  18      1          3        0   120
#> [283,]  63   1      1          0        1   155
#> [284,]  63  10      0         18        1   130
#> [285,]  67  11      0         11        0   150
#> [286,]  66  12      1         10        1   150
#> [287,]  65  15      1         12        1   150
#> [288,]  66  11      1          0        0   100
#> [289,]  65   4      1          2        1   145
#> [290,]  69  12      0         15        1   140
#> [291,]  65  11      1          6        0   130
#> [292,]  69  21      1         10        0   180
#> [293,]  66   9      1          8        0   130
#> [294,]  63   8      0        180        1   120
#> [295,]  68  14      1         13        1   140
#> [296,]  65   8      1          0        1    90
#> [297,]  66   3      0          3        1   138
#> [298,]  67   1      0        180        1   160
#> [299,]  67   7      1          4        1   130
#> [300,]  65  10      1          1        1   148
#> [301,]  67  12      1         12        0   160
#> [302,]  69   6      0         99        1   140
#> [303,]  64   4      0        179        0   160
#> [304,]  66   4      0        180        1   130
#> [305,]  70  15      1         12        1   132
#> [306,]  64  11      0         11        0   125
#> [307,]  64   4      0        180        1   140
#> [308,]  67   2      0         18        0   131
#> [309,]  66   7      1          5        1   131
#> [310,]  66   4      0        180        0   177
#> [311,]  69   4      1          3        1   150
#> [312,]  65  13      1         12        1   130
#> [313,]  69  17      1         10        0   140
#> [314,]  69   8      0         93        0   140
#> [315,]  66   6      0        180        0   140
#> [316,]  68  18      1          0        1   160
#> [317,]  65   6      0        101        1   115
#> [318,]  68   4      0          4        1   190
#> [319,]  71   3      0          5        0   112
#> [320,]  68   7      0        150        0   210
#> [321,]  66   9      1          3        1   151
#> [322,]  66   1      1          1        1   165
#> [323,]  70   4      1          0        1   180
#> [324,]  69   8      0        180        1   153
#> [325,]  70  14      0        171        0   166
#> [326,]  67   6      1          4        0   130
#> [327,]  67  14      1         13        0   130
#> [328,]  71   7      0          7        0   230
#> [329,]  66   2      0          2        1   228
#> [330,]  71   6      0         45        1   158
#> [331,]  69   3      0          3        1   130
#> [332,]  70  22      1         13        0   103
#> [333,]  67   1      0         36        1   104
#> [334,]  67   5      0          5        0   130
#> [335,]  69   6      1          4        1   174
#> [336,]  72   7      0          7        1   110
#> [337,]  69   8      1          7        1   108
#> [338,]  67   3      0        180        0   110
#> [339,]  66   2      1          1        0   123
#> [340,]  69  19      0        180        0   130
#> [341,]  69  11      1          0        1   120
#> [342,]  69   8      1          2        0   121
#> [343,]  67  13      1          9        0   130
#> [344,]  70   3      0        123        0   130
#> [345,]  70   9      0        180        1   142
#> [346,]  72   5      1          4        0   170
#> [347,]  67  22      1          1        1   140
#> [348,]  68   3      0         19        0   135
#> [349,]  67  12      1          8        0   120
#> [350,]  69   1      0          1        1   110
#> [351,]  67   1      0          1        1    60
#> [352,]  67   4      0         60        1   136
#> [353,]  69   5      0         76        0   120
#> [354,]  67   8      1          0        1   130
#> [355,]  72  13      1         11        1   195
#> [356,]  68  10      1          8        1   160
#> [357,]  70   7      0          7        0   102
#> [358,]  73  20      1          0        1   170
#> [359,]  71   6      0          9        0   120
#> [360,]  69  10      1          6        1   120
#> [361,]  70  11      0        180        1   210
#> [362,]  72  12      1         10        0   170
#> [363,]  67   5      1          0        1   147
#> [364,]  67   9      0        180        0   158
#> [365,]  73  13      0        152        1   130
#> [366,]  70   5      0        180        0   150
#> [367,]  67   4      1          1        0   134
#> [368,]  72   6      1          5        0   115
#> [369,]  68  23      0        180        1   220
#> [370,]  71   3      1          2        0   150
#> [371,]  68   4      1          3        0   210
#> [372,]  72   5      0         28        0   120
#> [373,]  71   5      0        180        0   191
#> [374,]  73   6      0        180        1   117
#> [375,]  69  16      1         10        1   140
#> [376,]  69   8      1          1        0   164
#> [377,]  72  16      1          1        1   130
#> [378,]  69   1      1          0        0   155
#> [379,]  73   6      1          0        1   270
#> [380,]  71   2      1          0        1   180
#> [381,]  73   7      0          7        1   140
#> [382,]  68  15      1         13        1   130
#> [383,]  70  13      1          9        0   100
#> [384,]  72   6      0        180        1   130
#> [385,]  73   0      0        180        1   161
#> [386,]  74   8      1          0        1    85
#> [387,]  69   2      1          0        1   110
#> [388,]  71   3      1          1        0   150
#> [389,]  74  20      0         20        1   180
#> [390,]  68   9      0        180        1   120
#> [391,]  71  20      1         10        0   140
#> [392,]  74   0      1          0        1    90
#> [393,]  73   3      1          0        1   136
#> [394,]  71  17      1         11        0   160
#> [395,]  71   8      1          7        0   149
#> [396,]  73  10      1          8        0   106
#> [397,]  74   4      0          4        0   120
#> [398,]  72   5      1          3        1   160
#> [399,]  70   3      0        180        1   154
#> [400,]  73   6      0        180        0   110
#> [401,]  72  15      1          0        1   150
#> [402,]  71   7      1          2        0   143
#> [403,]  74   3      0          3        1   150
#> [404,]  69   2      1          1        1    80
#> [405,]  70   4      1          0        1   140
#> [406,]  71  14      1         13        1   170
#> [407,]  74   7      1          0        1   117
#> [408,]  72  10      1          8        1   153
#> [409,]  69   7      0        180        1   144
#> [410,]  70   8      0          8        0   120
#> [411,]  75   1      0          1        0   133
#> [412,]  75   2      1          1        0   145
#> [413,]  73  10      1          9        1   146
#> [414,]  72  10      1          9        1   160
#> [415,]  73  10      1         10        1   120
#> [416,]  74  15      1          9        1   179
#> [417,]  71   2      0         10        1   112
#> [418,]  73   1      0          1        1    80
#> [419,]  75   9      1          7        0   140
#> [420,]  75  13      1          1        1   130
#> [421,]  71  11      1          8        0   110
#> [422,]  71   4      0          4        0   134
#> [423,]  72  15      1         12        1   120
#> [424,]  73  10      1          8        0   120
#> [425,]  70   7      1          4        0   184
#> [426,]  72   1      1          1        0   168
#> [427,]  73  10      0        180        0   162
#> [428,]  70   3      0          3        0   150
#> [429,]  73   5      1          3        1   112
#> [430,]  76  25      1         12        1   170
#> [431,]  72   2      0        180        0   120
#> [432,]  75   1      0        180        1   140
#> [433,]  72   4      1          0        1   197
#> [434,]  73   4      0        180        0   124
#> [435,]  74  34      1          8        1   233
#> [436,]  71  32      1         12        1   107
#> [437,]  72   5      0        180        0   154
#> [438,]  72   3      0        180        0   160
#> [439,]  77  11      0         11        1   150
#> [440,]  75   3      1          1        0   180
#> [441,]  73  15      0         15        1   160
#> [442,]  71  16      0        180        0   140
#> [443,]  73  10      1         10        0   124
#> [444,]  74   7      0        180        1   150
#> [445,]  74   3      0          3        1   128
#> [446,]  74  19      1          4        1   200
#> [447,]  73   6      0          6        1   114
#> [448,]  72   4      0         85        1   120
#> [449,]  72   4      1          3        0   160
#> [450,]  76  17      1          0        1   200
#> [451,]  73   4      1          3        1   125
#> [452,]  75   4      1          0        1   122
#> [453,]  75   0      0          0        1   130
#> [454,]  74   8      1          0        1   105
#> [455,]  76  13      1          8        1   148
#> [456,]  75   4      1          2        1   188
#> [457,]  74   6      0        180        0   160
#> [458,]  75   1      0          1        1   125
#> [459,]  74   2      0        180        0   111
#> [460,]  73   0      0        180        0   156
#> [461,]  72   5      0        180        0   120
#> [462,]  78  12      1         11        1   160
#> [463,]  76  44      1         10        0   105
#> [464,]  76   5      0        180        0   185
#> [465,]  73  33      1         12        1   175
#> [466,]  73  10      1          9        0   146
#> [467,]  77  12      0        180        0   130
#> [468,]  77   1      1          0        1    90
#> [469,]  78   5      1          0        1   170
#> [470,]  74   6      0         79        1   140
#> [471,]  75   3      1          1        1   171
#> [472,]  74   9      1          8        0   200
#> [473,]  79  10      1          8        0   190
#> [474,]  74   2      1          0        1   130
#> [475,]  78  18      0         18        1   144
#> [476,]  76  29      0         47        0    90
#> [477,]  73  11      1          2        1   110
#> [478,]  74   2      0        180        0   100
#> [479,]  78   7      0          7        1   133
#> [480,]  74  15      0        180        1   172
#> [481,]  76  13      1          1        1   170
#> [482,]  78  32      1          9        1   198
#> [483,]  79   6      0        180        0   170
#> [484,]  80  10      1          6        1   147
#> [485,]  78  13      1          5        0   130
#> [486,]  78  15      0        180        1   270
#> [487,]  80   8      0          8        1   120
#> [488,]  75  13      1          6        0   150
#> [489,]  74  10      1          8        0   135
#> [490,]  76   1      0          1        1    83
#> [491,]  79   4      0         80        0   145
#> [492,]  78  12      1          9        0   150
#> [493,]  75   4      1          0        0   212
#> [494,]  78  10      0        180        1   130
#> [495,]  76  11      1          0        0   120
#> [496,]  75  11      1          4        0   162
#> [497,]  77  24      0         24        1   160
#> [498,]  79   8      0         32        1   120
#> [499,]  80   6      0          6        1   150
#> [500,]  78   6      1          0        1   240
#> [501,]  78  11      1          1        1   140
#> [502,]  78  14      1          0        1   140
#> [503,]  81   1      0          1        0   130
#> [504,]  78  11      1          8        1   118
#> [505,]  76   4      0          4        1   160
#> [506,]  76  10      1          8        0   180
#> [507,]  76  12      1         10        1   127
#> [508,]  77   6      0          6        1   107
#> [509,]  75   2      1          1        1   204
#> [510,]  76   1      0          1        1    90
#> [511,]  78   7      1          0        1   110
#> [512,]  79   3      0          3        0   120
#> [513,]  77   7      0        180        1   170
#> [514,]  79   4      1          0        1   120
#> [515,]  81   1      0        180        0   120
#> [516,]  80  15      1         12        1   150
#> [517,]  77   9      1          4        0   141
#> [518,]  82   5      0          8        1   120
#> [519,]  80  40      1          0        1   138
#> [520,]  78   4      0         59        1   112
#> [521,]  76   7      0        161        0   151
#> [522,]  79  10      0         10        1   120
#> [523,]  80  15      1          0        1    90
#> [524,]  79  28      0        164        0   100
#> [525,]  78  32      0        180        1   130
#> [526,]  79   1      0         37        1   140
#> [527,]  81   2      0        175        0   172
#> [528,]  78   7      0          7        1   147
#> [529,]  77  13      1          0        1   190
#> [530,]  78  15      0         15        0   165
#> [531,]  76   1      0        166        0   131
#> [532,]  81   4      1          1        1   104
#> [533,]  78  20      1          0        1   109
#> [534,]  77  10      1          8        1   130
#> [535,]  82   3      1          1        1   144
#> [536,]  77   5      0         85        0   188
#> [537,]  79   6      0          6        0   152
#> [538,]  78   2      0        180        0   148
#> [539,]  80   5      0          5        1   130
#> [540,]  82   1      1          0        1    82
#> [541,]  77   4      0        180        1    98
#> [542,]  78  12      0        180        0   134
#> [543,]  82  21      1          2        0   155
#> [544,]  84  22      1         10        0   180
#> [545,]  82   5      0        180        0   110
#> [546,]  83   5      0        180        0   148
#> [547,]  79   7      1          6        0   130
#> [548,]  83   4      0        103        0    97
#> [549,]  81  11      1          8        0   160
#> [550,]  81   5      0        177        0    41
#> [551,]  80  11      1          8        0   170
#> [552,]  78  23      1         10        1   145
#> [553,]  79   4      0          4        1   183
#> [554,]  78   9      1          4        1   120
#> [555,]  82   8      1          1        0   128
#> [556,]  79   1      0        180        1   170
#> [557,]  81  15      0        180        1   140
#> [558,]  80   7      1          0        1   146
#> [559,]  84   5      1          1        1    85
#> [560,]  81  20      1          9        0   170
#> [561,]  83   8      0          8        0   115
#> [562,]  81  16      0         16        1   110
#> [563,]  80  11      1          8        0   110
#> [564,]  80   8      1          7        0   160
#> [565,]  79   0      1          0        1    96
#> [566,]  85   4      0        180        0    90
#> [567,]  81   2      1          1        0   198
#> [568,]  83   2      0          2        1   155
#> [569,]  82   6      0        128        1   100
#> [570,]  80   3      1          1        1   230
#> [571,]  84   5      0        180        1   203
#> [572,]  84   4      0          4        1    85
#> [573,]  84   1      0         38        1   205
#> [574,]  81   4      0         90        1   138
#> [575,]  85   3      1          2        1   160
#> [576,]  80   2      1          0        1   130
#> [577,]  79   4      0          4        1    60
#> [578,]  80   6      0         71        1   189
#> [579,]  82  19      0         19        0   120
#> [580,]  83   9      0        180        0   198
#> [581,]  79  14      1          0        0   110
#> [582,]  83   3      0        114        0    98
#> [583,]  81  14      1         12        1   128
#> [584,]  82   0      0          2        1   100
#> [585,]  81   4      0          4        0   175
#> [586,]  84  15      1         13        1   110
#> [587,]  81   1      0          1        1   145
#> [588,]  81  12      0         12        1   163
#> [589,]  82  16      1          8        0   103
#> [590,]  82   5      1          0        1   146
#> [591,]  86  12      0        180        1   120
#> [592,]  83   7      0        126        0   135
#> [593,]  86   8      0          8        1   132
#> [594,]  81  16      1          9        0   180
#> [595,]  86   3      0          3        1   140
#> [596,]  84   3      0        180        1   120
#> [597,]  81  13      0        180        0   152
#> [598,]  85   3      0          3        1   118
#> [599,]  81   2      1          0        1   118
#> [600,]  81   4      0        180        0   160
#> [601,]  82   1      0        180        1   193
#> [602,]  83   4      0          4        0   130
#> [603,]  86  12      1          0        1   132
#> [604,]  86   6      1          0        1   140
#> [605,]  84   3      1          2        0   125
#> [606,]  88  14      1          3        1   130
#> [607,]  83  13      1         12        0   170
#> [608,]  84   7      1          2        0   148
#> [609,]  87   2      0        180        0   113
#> [610,]  86   4      0         38        1   122
#> [611,]  82   4      0          4        0   130
#> [612,]  86  13      0        177        0   163
#> [613,]  85   3      0          3        1   113
#> [614,]  86   6      0          6        1   117
#> [615,]  84  13      0         62        1   100
#> [616,]  86   6      1          1        0   112
#> [617,]  88   4      0          4        1   115
#> [618,]  85  22      0         22        1   184
#> [619,]  83   9      0         65        1   150
#> [620,]  86   9      1          7        1   142
#> [621,]  86   6      0         46        0   173
#> [622,]  88   3      0        115        0   110
#> [623,]  87   8      0          8        1   157
#> [624,]  86  15      1          8        1   109
#> [625,]  87   1      0          1        0   170
#> [626,]  84   8      0        180        1   119
#> [627,]  84   2      0        110        1   174
#> [628,]  87  29      0         29        1    97
#> [629,]  87  15      1          9        1   138
#> [630,]  84   0      0        180        1   136
#> [631,]  89  10      0         46        1   170
#> [632,]  90  14      0         14        1   100
#> [633,]  88   1      0          1        0   135
#> [634,]  86   4      0        180        1   145
#> [635,]  87   6      1          0        0   125
#> [636,]  91  10      0        145        0   135
#> [637,]  88   8      0         50        1   154
#> [638,]  90  11      1         10        1   186
#> [639,]  87   6      0        126        1   168
#> [640,]  86   9      1          7        0   130
#> [641,]  90   4      1          0        0   121
#> [642,]  91   1      0          1        1    74
#> [643,]  87   5      0         36        1   150
#> [644,]  88   3      1          2        0   159
#> [645,]  91   3      0         33        1   137
#> [646,]  89  12      1          0        1   130
#> [647,]  89  52      0         52        1   130
#> [648,]  92   7      0          7        1   110
#> [649,]  89   4      0          4        1   159
#> [650,]  91   0      0          0        0     0
#> [651,]  89  14      0        180        1    84
#> [652,]  90  18      0        180        0   188
#> [653,]  91   4      1          0        1   120
#> [654,]  90  19      1         11        1   129
#> [655,]  94   6      0         50        0    78
#> [656,]  90   1      0          1        1   118
#> [657,]  93   8      0        179        1   110
#> [658,]  91   1      0        180        0   158
#> [659,]  90  16      0         16        1   106
#> [660,]  90   3      0         67        0   162
#> [661,]  96   3      0         12        1    97
#> [662,]  95   8      1          5        1   150
#> [663,]  94   3      0         26        1   144
#> [664,]  91  12      0         53        1   212
#> [665,]  91   7      0          7        0   135
#> [666,]  93   0      1          0        1   122
#> [667,]  92   5      0         69        0   139
#> [668,]  92   2      0          2        0   112
#> [669,]  93   4      0        180        1   135
#> [670,]  96   3      1          0        1   104
#> 
#> $y
#>   [1] 180.0+   5.0+   5.0+   5.0+ 180.0+   2.0+ 180.0+ 115.0  180.0+ 180.0+
#>  [11]  12.0    5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 180.0+   2.0+
#>  [21] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 180.0+ 155.0+
#>  [31] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [41] 180.0+ 150.0  180.0+ 180.0+ 180.0+ 180.0+   6.0+ 180.0+ 180.0+ 180.0+
#>  [51] 180.0+  73.0  161.0+ 180.0+ 180.0+ 180.0+   5.0+ 180.0+ 180.0+ 180.0+
#>  [61] 177.0+ 180.0+ 180.0+ 180.0+ 180.0+  10.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [71] 180.0+   1.0  179.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [81] 180.0+ 180.0+ 180.0+   7.0   36.0  180.0+   4.0+ 180.0+  77.0  180.0+
#>  [91] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  85.0  180.0+ 180.0+
#> [101] 166.0+  99.0  180.0+  16.0+ 152.0+   7.0+   6.0+ 180.0+  13.0+  28.0 
#> [111]   6.0+   1.0    9.0+ 180.0+ 180.0+ 180.0+   2.0  180.0+   7.0+ 180.0+
#> [121] 180.0+ 180.0+ 180.0+ 180.0+  16.0+  16.0   15.0+ 180.0+ 180.0+  12.0+
#> [131]   8.0    3.0+ 180.0+ 180.0+ 180.0+ 140.0  180.0+   1.0  165.0  180.0+
#> [141] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 171.0+  15.0  166.0+   5.0+ 180.0+
#> [151] 180.0+ 180.0+   5.0+ 180.0+   1.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [161] 180.0+ 180.0+ 180.0+ 180.0+  64.0  180.0+ 180.0+ 180.0+ 180.0+   0.5 
#> [171] 180.0+ 161.0+ 171.0+   3.0    1.0  180.0+ 180.0+  10.0+  17.0   45.0 
#> [181]   3.0+  19.0  180.0+ 180.0+   9.0+ 180.0+ 172.0+ 172.0+   8.0  180.0+
#> [191] 180.0+   1.0+  15.0  180.0+ 180.0+ 180.0+  77.0    8.0+ 180.0+ 170.0 
#> [201]  94.0  169.0    7.0+   6.0  180.0+ 180.0+ 180.0+ 180.0+   3.0+  30.0 
#> [211] 180.0+   5.0   18.0    5.0+ 180.0+ 180.0+   1.0  180.0+ 180.0+ 180.0+
#> [221] 180.0+ 180.0+   4.0+ 180.0+   9.0   22.0  180.0+ 180.0+  84.0  180.0+
#> [231] 180.0+   1.0    1.0  180.0+ 167.0  180.0+ 180.0+ 180.0+ 180.0+  14.0+
#> [241]  36.0  180.0+   3.0+  88.0  180.0+  12.0  180.0+ 180.0+ 180.0+   0.5 
#> [251] 180.0+ 180.0+ 180.0+  12.0  180.0+   3.0  180.0+ 180.0+   2.0+  18.0+
#> [261] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   3.0+  15.0  180.0+ 179.0+ 166.0+
#> [271]  14.0+   0.5+ 180.0+   3.0+ 180.0+   7.0+   5.0    1.0  180.0+ 180.0+
#> [281] 180.0+ 123.0+   1.0+  18.0   11.0+  80.0   15.0+ 180.0+   4.0+  15.0 
#> [291] 180.0+ 174.0+ 180.0+ 180.0+ 180.0+   8.0+   3.0  180.0+ 180.0+ 180.0+
#> [301]  12.0   99.0  179.0+ 180.0+ 180.0+  11.0+ 180.0+  18.0    7.0+ 180.0+
#> [311] 152.0+ 180.0+ 180.0+  93.0  180.0+  18.0+ 101.0    4.0    5.0  150.0 
#> [321] 180.0+   1.0  180.0+ 180.0+ 171.0    6.0  180.0+   7.0+   2.0   45.0 
#> [331]   3.0+ 180.0+  36.0    5.0+  97.0    7.0    8.0+ 180.0+   2.0+ 180.0+
#> [341] 180.0+   8.0+  13.0+ 123.0  180.0+ 180.0+  51.0   19.0  180.0+   1.0 
#> [351]   1.0   60.0   76.0  180.0+ 132.0   10.0+   7.0+ 124.0    9.0  180.0+
#> [361] 180.0+  12.0  180.0+ 180.0+ 152.0  180.0+  76.0  180.0+ 180.0+ 180.0+
#> [371] 180.0+  28.0  180.0+ 180.0+  16.0+ 180.0+  16.0+ 180.0+   6.0  180.0+
#> [381]   7.0+  15.0   13.0+ 180.0+ 180.0+ 180.0+   2.0    3.0+  20.0  180.0+
#> [391]  20.0    0.5  180.0+ 180.0+   8.0   87.0    4.0+ 180.0+ 180.0+ 180.0+
#> [401] 180.0+ 180.0+   3.0    2.0  180.0+  14.0+ 180.0+  10.0+ 180.0+   8.0+
#> [411]   1.0  180.0+ 180.0+ 159.0   15.0  180.0+  10.0    1.0  180.0+  13.0 
#> [421] 180.0+   4.0+ 180.0+  10.0  104.0+   1.0  180.0+   3.0+   5.0  180.0+
#> [431] 180.0+ 180.0+ 180.0+ 180.0+  34.0  177.0+ 180.0+ 180.0+  11.0+ 180.0+
#> [441]  15.0+ 180.0+  10.0  180.0+   3.0  180.0+   6.0   85.0  180.0+  17.0+
#> [451] 180.0+   4.0    0.5  180.0+ 180.0+  46.0  180.0+   1.0  180.0+ 180.0+
#> [461] 180.0+  12.0  180.0+ 180.0+  33.0  180.0+ 180.0+   1.0  180.0+  79.0 
#> [471]   3.0  168.0+ 180.0+ 176.0+  18.0   47.0   11.0  180.0+   7.0  180.0+
#> [481] 180.0+  32.0  180.0+  10.0  172.0  180.0+   8.0  180.0+ 180.0+   1.0 
#> [491]  80.0  180.0+   4.0+ 180.0+  11.0  152.0+  24.0   32.0    6.0  180.0+
#> [501] 180.0+ 180.0+   1.0   11.0    4.0   10.0+ 180.0+   6.0    2.0+   1.0 
#> [511]  43.0    3.0  180.0+ 138.0  180.0+ 180.0+  71.0    8.0   40.0   59.0 
#> [521] 161.0   10.0+ 180.0+ 164.0  180.0+  37.0  175.0+   7.0+  22.0   15.0+
#> [531] 166.0+  71.0   20.0+  10.0  180.0+  85.0    6.0+ 180.0+   5.0    1.0 
#> [541] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 103.0  180.0+ 177.0+
#> [551] 169.0   70.0    4.0  180.0+ 180.0+ 180.0+ 180.0+   7.0+ 180.0+  20.0 
#> [561]   8.0+  16.0  180.0+ 180.0+   0.5  180.0+ 180.0+   2.0  128.0    3.0+
#> [571] 180.0+   4.0   38.0   90.0  180.0+ 180.0+   4.0   71.0   19.0  180.0+
#> [581] 180.0+ 114.0  180.0+   2.0    4.0+ 180.0+   1.0   12.0   16.0+   5.0+
#> [591] 180.0+ 126.0    8.0  180.0+   3.0  180.0+ 180.0+   3.0+ 180.0+ 180.0+
#> [601] 180.0+   4.0+ 180.0+   6.0  180.0+  14.0   13.0  180.0+ 180.0+  38.0 
#> [611]   4.0  177.0    3.0+   6.0+  62.0    6.0+   4.0   22.0   65.0   11.0 
#> [621]  46.0  115.0    8.0+ 180.0+   1.0+ 180.0+ 110.0   29.0  180.0+ 180.0+
#> [631]  46.0   14.0    1.0+ 180.0+  25.0  145.0   50.0   11.0  126.0  180.0+
#> [641]   4.0    1.0   36.0   75.0   33.0  180.0+  52.0    7.0    4.0    0.5 
#> [651] 180.0+ 180.0+   4.0  180.0+  50.0    1.0+ 179.0+ 180.0+  16.0   67.0 
#> [661]  12.0    8.0   26.0   53.0    7.0+   0.5   69.0    2.0  180.0+   3.0 
#> 
#> $weights
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8731568