Cross-Validated GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::cv.glmnet() from package glmnet.

Details

This learner returns two prediction types:

lp: a vector of linear predictors (relative risk scores), one per observation. Calculated using glmnet::predict.cv.glmnet().

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.cv.glmnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used.

Custom mlr3 parameters

family is set to "cox" and cannot be changed.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.cv_glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

foldid	untyped	NULL		-

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nfolds	integer	10		\([3, \infty)\)

nlambda	integer	100		\([1, \infty)\)

offset	untyped	NULL		-

newoffset	untyped	-		-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	lambda.1se		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

standardize.response	logical	FALSE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.gaussian	character	-	covariance, naive	-

type.logistic	character	Newton	Newton, modified.Newton	-

type.measure	character	deviance	deviance, C	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvCVGlmnet

Methods

Public methods

LearnerSurvCVGlmnet$new()

LearnerSurvCVGlmnet$selected_features()

LearnerSurvCVGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvCVGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvCVGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvCVGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.cv_glmnet")
print(learner)
#> <LearnerSurvCVGlmnet:surv.cv_glmnet>: Regularized Generalized Linear Model
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3proba, mlr3extralearners, glmnet
#> * Predict Types:  [crank], distr, lp
#> * Feature Types: logical, integer, numeric
#> * Properties: selected_features, weights

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#> Measure: Partial Likelihood Deviance 
#> 
#>      Lambda Index Measure     SE Nonzero
#> min 0.00375    43   8.479 0.1151       6
#> 1se 0.03494    19   8.591 0.1324       6
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  28   9      0        180        1   107
#>   [2,]  33   2      0          2        0   150
#>   [3,]  35  10      1          9        0   106
#>   [4,]  34   5      0          5        0   120
#>   [5,]  35   2      0        180        0   121
#>   [6,]  35   2      1          1        1   112
#>   [7,]  37   9      0        180        1   151
#>   [8,]  38  13      1          0        1   161
#>   [9,]  38   2      0        115        0   150
#>  [10,]  36   1      0        180        1   155
#>  [11,]  36   5      1          0        1   115
#>  [12,]  33   6      1          1        1   115
#>  [13,]  38  16      1         10        0   160
#>  [14,]  37   1      1          0        1   146
#>  [15,]  40   2      1          1        1   148
#>  [16,]  38   5      1          3        0   125
#>  [17,]  42   2      0          2        0   140
#>  [18,]  43   3      1          0        1   100
#>  [19,]  40   1      1          0        1   145
#>  [20,]  43   4      1          0        1   130
#>  [21,]  42  15      1         13        1   125
#>  [22,]  40   3      1          1        0   170
#>  [23,]  42  12      1         10        1   170
#>  [24,]  42   2      0        180        1   124
#>  [25,]  44   3      0        180        0   141
#>  [26,]  41  13      1          1        0   140
#>  [27,]  45   6      0        180        1   170
#>  [28,]  44   2      1          1        1   150
#>  [29,]  43   2      0        180        1   140
#>  [30,]  45   2      0        180        1   140
#>  [31,]  46  15      0        180        0   120
#>  [32,]  46   2      1          1        0   126
#>  [33,]  47   4      1          3        0   118
#>  [34,]  48  15      0        180        1   160
#>  [35,]  44   3      1          0        1   180
#>  [36,]  43  29      0        180        1   180
#>  [37,]  45   4      1          0        0   124
#>  [38,]  43  10      0        180        0   185
#>  [39,]  46  13      1         10        0   100
#>  [40,]  44   0      1          0        1    96
#>  [41,]  47   4      1          3        1   160
#>  [42,]  43   3      1          0        1   124
#>  [43,]  45   8      1          0        1   117
#>  [44,]  49   5      0         73        1   136
#>  [45,]  45   5      0          5        0   141
#>  [46,]  46   2      1          1        1   122
#>  [47,]  46   6      1          0        1   100
#>  [48,]  44   4      1          0        1   114
#>  [49,]  47   2      0        180        0   108
#>  [50,]  44   9      1          8        1   135
#>  [51,]  46   5      1          3        0   130
#>  [52,]  46   4      0        180        1   121
#>  [53,]  44   2      0        180        0   142
#>  [54,]  46  15      0        180        1   120
#>  [55,]  45   9      1          0        1   145
#>  [56,]  47   3      1          1        1   120
#>  [57,]  48  12      1         11        0   200
#>  [58,]  47   5      1          3        1   130
#>  [59,]  47   9      1          6        0   170
#>  [60,]  46   3      1          0        1   119
#>  [61,]  47  10      0         10        1   140
#>  [62,]  50   1      1          0        1   129
#>  [63,]  47   7      0        180        0   145
#>  [64,]  50   4      1          1        0   125
#>  [65,]  46   3      1          1        1   140
#>  [66,]  50   7      0        180        1   110
#>  [67,]  51   1      0          1        1   145
#>  [68,]  47   2      0        180        0   150
#>  [69,]  49  23      0        179        1   112
#>  [70,]  46   6      1          0        1   156
#>  [71,]  52   2      0        180        1   170
#>  [72,]  50   7      1          0        1    92
#>  [73,]  50   9      0        180        0   130
#>  [74,]  49   7      1          4        1    90
#>  [75,]  47   8      0        180        0   160
#>  [76,]  47   6      0        180        1   162
#>  [77,]  52   2      0        180        0   155
#>  [78,]  46   3      0        180        1   120
#>  [79,]  50   4      1          1        0   150
#>  [80,]  53   8      0         36        1   160
#>  [81,]  48  17      1         10        0   111
#>  [82,]  49  15      0        180        1   160
#>  [83,]  53   5      0        180        1   140
#>  [84,]  54  17      1         12        1   102
#>  [85,]  53   7      1          0        0   199
#>  [86,]  54   6      1          3        0   129
#>  [87,]  51   3      1          1        0   140
#>  [88,]  50   2      0          5        1   106
#>  [89,]  50  10      1          6        0   122
#>  [90,]  50  14      1         13        0   170
#>  [91,]  48   3      1          2        0   150
#>  [92,]  51  25      1          1        0   202
#>  [93,]  49   5      1          2        1   150
#>  [94,]  52  14      1          7        1   200
#>  [95,]  48  11      1         10        0   120
#>  [96,]  53   4      1          0        1   156
#>  [97,]  51  13      0         99        1   160
#>  [98,]  49  16      0         16        0   125
#>  [99,]  55   3      1          1        0   150
#> [100,]  52   7      1          2        0   154
#> [101,]  55   6      1          2        1   114
#> [102,]  54   9      1          1        0   130
#> [103,]  55   4      1          2        0   150
#> [104,]  51  13      1         11        0   145
#> [105,]  50   5      1          4        1   150
#> [106,]  54   4      1          0        1   121
#> [107,]  50   3      0        174        1   153
#> [108,]  49   6      1          0        1   130
#> [109,]  49   1      0          1        1   110
#> [110,]  50   7      1          1        0   156
#> [111,]  53   8      1          0        1   130
#> [112,]  50   7      1          0        1   127
#> [113,]  56   4      1          1        1   130
#> [114,]  52   5      0        175        1   117
#> [115,]  55   1      0        180        0   127
#> [116,]  55   2      0          2        0   145
#> [117,]  54   1      0        180        0   162
#> [118,]  54   7      1          0        1   100
#> [119,]  56   2      0        180        0   132
#> [120,]  55   5      1          4        1   120
#> [121,]  52   8      0        180        0   119
#> [122,]  52  16      0         16        0   152
#> [123,]  53  10      1          9        0   172
#> [124,]  52  16      1         14        0   170
#> [125,]  53  15      0         15        1    90
#> [126,]  53   4      0        180        1   150
#> [127,]  55   6      0        180        1   100
#> [128,]  55   6      1          5        1   138
#> [129,]  54  12      1          0        1   190
#> [130,]  56   3      0          8        1   139
#> [131,]  55   1      0          2        0   130
#> [132,]  57   3      0          3        0   120
#> [133,]  54   7      1          2        0   129
#> [134,]  54   2      1          1        0   135
#> [135,]  52   9      1          3        0   170
#> [136,]  57   5      1          3        1   138
#> [137,]  57   1      0        180        1   156
#> [138,]  56   4      1          0        1   140
#> [139,]  56  14      1         11        0   130
#> [140,]  55   3      1          1        1   156
#> [141,]  57   0      0          0        1   150
#> [142,]  59   3      1          1        0   172
#> [143,]  57   4      0        180        1   119
#> [144,]  53  15      1         10        1   130
#> [145,]  54  17      1          8        1   227
#> [146,]  55   9      1          2        1   147
#> [147,]  56   5      0          5        1   150
#> [148,]  57   4      1          2        1   185
#> [149,]  53   7      1          0        1   120
#> [150,]  55   3      1          2        0   140
#> [151,]  55   5      0          5        1   131
#> [152,]  54   7      1          0        1   141
#> [153,]  56   4      0          4        0   164
#> [154,]  58   9      1          0        1   180
#> [155,]  58   1      1          1        1   200
#> [156,]  55   2      0          2        0   106
#> [157,]  59   9      1          1        1   125
#> [158,]  57   1      0        180        0   148
#> [159,]  60  11      1          9        0   106
#> [160,]  58   4      1          0        1   160
#> [161,]  57   2      0          2        1   120
#> [162,]  58  11      1          9        1   124
#> [163,]  55   5      1          0        1   160
#> [164,]  57  10      1          9        0   103
#> [165,]  59   6      1          0        1   140
#> [166,]  59   4      1          0        1   152
#> [167,]  58  26      1          0        1   189
#> [168,]  58   4      1          3        0   120
#> [169,]  60   0      1          0        1    80
#> [170,]  59   2      1          1        0   140
#> [171,]  58   8      0        161        1   140
#> [172,]  58  14      1          6        0   190
#> [173,]  61   4      1          3        0   151
#> [174,]  61   9      1          8        0   150
#> [175,]  61   3      1          2        1   102
#> [176,]  58   1      0          1        1   100
#> [177,]  57  13      1         10        0   110
#> [178,]  57   2      1          0        1   116
#> [179,]  57  11      0        180        1   150
#> [180,]  57   3      1          2        0   120
#> [181,]  58  19      1         13        1   140
#> [182,]  56  13      1          6        1   158
#> [183,]  59   9      1          0        1    80
#> [184,]  55   9      1          7        1   135
#> [185,]  56   8      1          8        0   120
#> [186,]  61  13      1         12        1   130
#> [187,]  59  11      1          8        1   190
#> [188,]  59   5      1          2        0   182
#> [189,]  58   5      1          1        1   135
#> [190,]  59  10      0        180        0   160
#> [191,]  61   8      0         77        0   120
#> [192,]  61  13      0         13        0   210
#> [193,]  62  10      1          0        1   153
#> [194,]  62   7      1          2        1   180
#> [195,]  57   3      1          0        0   100
#> [196,]  61  28      1          7        0   133
#> [197,]  58   8      1          3        1   150
#> [198,]  60   7      0          7        0   147
#> [199,]  61   6      0          6        0   134
#> [200,]  62   4      1          0        0   160
#> [201,]  58   3      1          0        1   146
#> [202,]  58   2      0         30        0   202
#> [203,]  59   1      0        180        0   155
#> [204,]  59  16      1          9        1   133
#> [205,]  63   6      0         28        1   120
#> [206,]  61   5      0          5        1   110
#> [207,]  57  18      1          9        1    93
#> [208,]  61   5      0          5        1   160
#> [209,]  57   2      1          1        0   159
#> [210,]  62  17      1         10        1   180
#> [211,]  62   1      1          0        1   172
#> [212,]  63   3      1          1        0   180
#> [213,]  63   1      0        180        1   130
#> [214,]  63   4      1          3        0   222
#> [215,]  62   3      0        180        1   105
#> [216,]  64   4      0        180        0   130
#> [217,]  59   8      0        180        1   140
#> [218,]  61   9      1          9        1   150
#> [219,]  58   9      1          9        0   110
#> [220,]  62   7      0          7        0   150
#> [221,]  59   1      0         22        1   162
#> [222,]  60   7      1          5        1   141
#> [223,]  60   7      0          7        0   140
#> [224,]  65  13      0        180        1   100
#> [225,]  63   1      0          1        0   162
#> [226,]  63   1      0          1        0   130
#> [227,]  61  15      1         13        0   170
#> [228,]  59   4      0          4        0   149
#> [229,]  62   6      0          6        0   120
#> [230,]  63  12      1         10        0   200
#> [231,]  59  10      0        180        1   130
#> [232,]  61   6      1          1        1   117
#> [233,]  64  12      1         11        0   160
#> [234,]  66   1      1          0        1   120
#> [235,]  65  36      1         11        0   140
#> [236,]  63   4      1          3        0   162
#> [237,]  66   3      1          1        0   127
#> [238,]  61  10      1          2        1   194
#> [239,]  64  32      1          9        1   160
#> [240,]  63  12      1          9        0   114
#> [241,]  63   7      0        180        0   120
#> [242,]  66   5      1          0        1   110
#> [243,]  65   8      1          0        0   168
#> [244,]  65  10      1          8        1   120
#> [245,]  64   0      0          0        1    90
#> [246,]  60   6      0        180        0   130
#> [247,]  61  12      1         11        0   154
#> [248,]  64   9      0        180        0   150
#> [249,]  61   4      0        180        1   113
#> [250,]  65   3      0        180        1   190
#> [251,]  64   7      0        180        1   120
#> [252,]  66   6      1          1        1   130
#> [253,]  63  12      0         12        1   150
#> [254,]  65   6      0          9        0   112
#> [255,]  66  16      1         11        1   169
#> [256,]  62   9      0        180        0   145
#> [257,]  61  14      1          5        0   140
#> [258,]  63   9      1          8        1   160
#> [259,]  63   3      1          2        0   120
#> [260,]  64  19      1          8        1   160
#> [261,]  65   8      1          0        1   140
#> [262,]  65  15      1         11        1   160
#> [263,]  68   5      1          4        1   150
#> [264,]  64  13      1         12        1   150
#> [265,]  66   7      1          0        1   115
#> [266,]  64   0      0          0        1   148
#> [267,]  67   4      1          3        0   130
#> [268,]  66   3      1          0        1   135
#> [269,]  65   2      1          1        1   170
#> [270,]  68   1      0        180        1   166
#> [271,]  63   7      1          0        0   162
#> [272,]  67   8      1          1        1   130
#> [273,]  64   1      0          1        1   120
#> [274,]  68  18      0        180        1   260
#> [275,]  65  17      1         14        1   100
#> [276,]  63   8      1          1        1   162
#> [277,]  65  18      1          3        0   120
#> [278,]  63   1      1          0        1   155
#> [279,]  63  10      0         18        1   130
#> [280,]  68  11      0        180        0   160
#> [281,]  68  14      0         79        0   172
#> [282,]  66  11      1          0        0   100
#> [283,]  65   4      1          2        1   145
#> [284,]  66  15      1         13        1   160
#> [285,]  63   2      0        180        0   150
#> [286,]  69   6      0        180        1   100
#> [287,]  66   9      1          8        0   130
#> [288,]  68  14      1         13        1   140
#> [289,]  66   3      0          3        1   138
#> [290,]  69   1      1          0        0   170
#> [291,]  68  10      1         10        1   150
#> [292,]  65   1      1          0        0   133
#> [293,]  67   7      1          4        1   130
#> [294,]  67   2      0        180        0   184
#> [295,]  65   6      0          6        0    80
#> [296,]  66  19      1         12        1   150
#> [297,]  67  12      1         12        0   160
#> [298,]  69   6      0         99        1   140
#> [299,]  65   4      1          1        0   130
#> [300,]  70  15      1         12        1   132
#> [301,]  64   0      1          0        1   118
#> [302,]  66   7      1          5        1   131
#> [303,]  68   4      1          0        1   160
#> [304,]  69   4      1          3        1   150
#> [305,]  65  13      1         12        1   130
#> [306,]  69  17      1         10        0   140
#> [307,]  69   8      0         93        0   140
#> [308,]  66   6      0        180        0   140
#> [309,]  65   1      0          1        1   120
#> [310,]  68  18      1          0        1   160
#> [311,]  65   6      0        101        1   115
#> [312,]  68   4      0          4        1   190
#> [313,]  71   3      0          5        0   112
#> [314,]  70   7      1          0        1   190
#> [315,]  68   7      0        150        0   210
#> [316,]  71  20      1          0        1   160
#> [317,]  67   2      0        180        0   128
#> [318,]  66   9      1          3        1   151
#> [319,]  69   8      0        180        1   153
#> [320,]  67   6      1          4        0   130
#> [321,]  68  18      1         14        1   170
#> [322,]  69   0      0          0        1   148
#> [323,]  65   2      0        180        0   130
#> [324,]  68   7      1          0        1   150
#> [325,]  65  14      1         13        1   150
#> [326,]  69   8      0        180        1   180
#> [327,]  71   7      0          7        0   230
#> [328,]  66   2      0          2        1   228
#> [329,]  71   6      0         45        1   158
#> [330,]  69   5      0          5        1   142
#> [331,]  70  22      1         13        0   103
#> [332,]  67   5      0          5        0   130
#> [333,]  72   3      1          0        1   132
#> [334,]  72   7      0          7        1   110
#> [335,]  69   8      1          7        1   108
#> [336,]  67   3      0        180        0   110
#> [337,]  69  19      0        180        0   130
#> [338,]  67  14      0        172        1   140
#> [339,]  69   4      1          3        0   132
#> [340,]  69   8      1          2        0   121
#> [341,]  67  13      1          9        0   130
#> [342,]  70   9      0        180        1   142
#> [343,]  72   5      1          4        0   170
#> [344,]  68   3      0         19        0   135
#> [345,]  67  12      1          8        0   120
#> [346,]  69   1      0          1        1   110
#> [347,]  67   1      0          1        1    60
#> [348,]  67   4      0         60        1   136
#> [349,]  72  13      1         11        1   195
#> [350,]  68  10      1          8        1   160
#> [351,]  66  24      1         13        0   130
#> [352,]  72  30      1          0        1   145
#> [353,]  70   7      0          7        0   102
#> [354,]  68   7      1          2        0   135
#> [355,]  71   6      0          9        0   120
#> [356,]  69  10      1          6        1   120
#> [357,]  70  11      0        180        1   210
#> [358,]  72  12      1         10        0   170
#> [359,]  67   8      0        180        1   170
#> [360,]  67   5      1          0        1   147
#> [361,]  70   5      0        180        0   150
#> [362,]  72   2      0          2        1   100
#> [363,]  67   4      1          1        0   134
#> [364,]  72   6      1          5        0   115
#> [365,]  71   1      0        173        1   188
#> [366,]  70   3      0        180        0   121
#> [367,]  68   4      1          3        0   210
#> [368,]  72   5      0         28        0   120
#> [369,]  71   5      0        180        0   191
#> [370,]  69  16      1         10        1   140
#> [371,]  68   7      0        180        1   130
#> [372,]  72  16      1          1        1   130
#> [373,]  70   4      0        180        0   180
#> [374,]  69   1      1          0        0   155
#> [375,]  73   7      0          7        1   140
#> [376,]  68  15      1         13        1   130
#> [377,]  70   3      0          3        1   159
#> [378,]  70  13      1          9        0   100
#> [379,]  74   8      1          0        1    85
#> [380,]  69   2      1          0        1   110
#> [381,]  71   3      1          1        0   150
#> [382,]  74  20      0         20        1   180
#> [383,]  68   9      0        180        1   120
#> [384,]  71  20      1         10        0   140
#> [385,]  74   0      1          0        1    90
#> [386,]  73   3      1          0        1   136
#> [387,]  70   5      1          0        1   190
#> [388,]  71  17      1         11        0   160
#> [389,]  71   3      1          2        1   190
#> [390,]  69  12      1          1        1   149
#> [391,]  73   4      0         58        1   160
#> [392,]  72   5      1          3        1   160
#> [393,]  70   3      0        180        1   154
#> [394,]  73   6      0        180        0   110
#> [395,]  72  15      1          0        1   150
#> [396,]  71   7      1          2        0   143
#> [397,]  72   8      1          0        1   140
#> [398,]  74   3      0          3        1   150
#> [399,]  71  13      1          8        0   121
#> [400,]  69   2      1          1        1    80
#> [401,]  70   4      1          0        1   140
#> [402,]  71  14      1         13        1   170
#> [403,]  72  10      1          8        1   153
#> [404,]  71  10      1          9        1   120
#> [405,]  72  10      1          9        1   160
#> [406,]  71   2      0         10        1   112
#> [407,]  73   1      0          1        1    80
#> [408,]  75  13      1          1        1   130
#> [409,]  71  11      1          8        0   110
#> [410,]  72  15      1         12        1   120
#> [411,]  73  10      1          8        0   120
#> [412,]  70   7      1          4        0   184
#> [413,]  72   1      1          1        0   168
#> [414,]  73  10      0        180        0   162
#> [415,]  72  11      0         11        1   140
#> [416,]  73   5      1          3        1   112
#> [417,]  76  25      1         12        1   170
#> [418,]  73  12      1         12        1   140
#> [419,]  72   2      0        180        0   120
#> [420,]  72   4      1          0        1   197
#> [421,]  71   3      1          0        0   144
#> [422,]  73   5      0        180        0   126
#> [423,]  73   4      0        180        0   124
#> [424,]  74  34      1          8        1   233
#> [425,]  76   3      1          0        1   120
#> [426,]  71  32      1         12        1   107
#> [427,]  72   5      0        180        0   154
#> [428,]  77  11      0         11        1   150
#> [429,]  77   4      0          4        0   185
#> [430,]  73  15      0         15        1   160
#> [431,]  71  16      0        180        0   140
#> [432,]  73  10      1         10        0   124
#> [433,]  74   7      0        180        1   150
#> [434,]  76   1      0        180        0   114
#> [435,]  76   8      1          0        1   141
#> [436,]  74  19      1          4        1   200
#> [437,]  73   6      0          6        1   114
#> [438,]  75  23      1         14        1   110
#> [439,]  74   2      0        180        0   190
#> [440,]  72   4      0         85        1   120
#> [441,]  72   4      1          3        0   160
#> [442,]  76  17      1          0        1   200
#> [443,]  75   7      0          7        0   190
#> [444,]  75   0      0          0        1   130
#> [445,]  74   8      1          0        1   105
#> [446,]  76  13      1          8        1   148
#> [447,]  75   4      1          2        1   188
#> [448,]  75   1      0          1        1   125
#> [449,]  73   0      0        180        0   156
#> [450,]  72   5      0        180        0   120
#> [451,]  78  12      1         11        1   160
#> [452,]  76  44      1         10        0   105
#> [453,]  76   5      0        180        0   185
#> [454,]  74  10      1          0        1   135
#> [455,]  74   8      1          8        1   170
#> [456,]  75   9      0        180        1   140
#> [457,]  77   5      1          0        0   123
#> [458,]  77  12      1          9        1   100
#> [459,]  73  10      1          9        0   146
#> [460,]  77  12      0        180        0   130
#> [461,]  78   5      1          0        1   170
#> [462,]  73   7      1          0        0   174
#> [463,]  74   6      0         79        1   140
#> [464,]  75   3      1          1        1   171
#> [465,]  74   9      1          8        0   200
#> [466,]  75   6      0        180        0   150
#> [467,]  79  10      1          8        0   190
#> [468,]  74   2      1          0        1   130
#> [469,]  77   3      0        180        0   110
#> [470,]  76  29      0         47        0    90
#> [471,]  73   8      1          1        1   162
#> [472,]  73  11      1          2        1   110
#> [473,]  78   7      0          7        1   133
#> [474,]  74  15      0        180        1   172
#> [475,]  78   8      1          6        1   110
#> [476,]  76  13      1          1        1   170
#> [477,]  78  32      1          9        1   198
#> [478,]  79   6      0        180        0   170
#> [479,]  80  10      1          6        1   147
#> [480,]  78   0      0        180        1   212
#> [481,]  78  13      1          5        0   130
#> [482,]  75   5      0        119        1   150
#> [483,]  75  12      1          1        1   120
#> [484,]  78  15      0        180        1   270
#> [485,]  80   8      0          8        1   120
#> [486,]  76   1      0          1        1    83
#> [487,]  79   4      0         80        0   145
#> [488,]  78  12      1          9        0   150
#> [489,]  78   2      1          1        0   130
#> [490,]  75   4      1          0        0   212
#> [491,]  75  11      1          4        0   162
#> [492,]  75   3      0          3        0     0
#> [493,]  80   6      0          6        1   150
#> [494,]  78   6      1          0        1   240
#> [495,]  78  11      1          1        1   140
#> [496,]  79  11      0        180        0   160
#> [497,]  79   2      1          0        1   121
#> [498,]  78  14      1          0        1   140
#> [499,]  79   4      0          4        1   125
#> [500,]  76  10      1          8        0   180
#> [501,]  76  12      1         10        1   127
#> [502,]  77   6      0          6        1   107
#> [503,]  80   3      1          0        1   120
#> [504,]  75   2      1          1        1   204
#> [505,]  78  11      0        180        1   135
#> [506,]  76   1      0          1        1   140
#> [507,]  77  31      1          3        1   161
#> [508,]  78   7      1          0        1   110
#> [509,]  77   6      0          6        1   144
#> [510,]  79   4      1          0        1   120
#> [511,]  80  15      1         12        1   150
#> [512,]  77   9      1          4        0   141
#> [513,]  78   4      0         59        1   112
#> [514,]  80  17      1         12        0   100
#> [515,]  76   7      0        161        0   151
#> [516,]  80  15      1          0        1    90
#> [517,]  79  28      0        164        0   100
#> [518,]  80   6      0        173        1   160
#> [519,]  78  32      0        180        1   130
#> [520,]  79   1      0         37        1   140
#> [521,]  81   3      0        180        0   184
#> [522,]  80   5      1          1        1   108
#> [523,]  78   4      0        180        0   175
#> [524,]  79   3      0          3        1   101
#> [525,]  78  26      1          5        0   194
#> [526,]  78  20      1          0        1   109
#> [527,]  77  10      1          8        1   130
#> [528,]  80   2      1          1        0   168
#> [529,]  79   6      0          6        0   152
#> [530,]  80   6      1          0        1   119
#> [531,]  80   5      0          5        1   130
#> [532,]  79  10      0        180        1   150
#> [533,]  77   4      0        180        1    98
#> [534,]  81   1      0        108        0   129
#> [535,]  79   1      0        125        0   193
#> [536,]  82  21      1          2        0   155
#> [537,]  84  22      1         10        0   180
#> [538,]  82   5      0        180        0   110
#> [539,]  79   7      1          6        0   130
#> [540,]  83   4      0        103        0    97
#> [541,]  81  11      1          8        0   160
#> [542,]  81   5      0        177        0    41
#> [543,]  80  11      1          8        0   170
#> [544,]  78  23      1         10        1   145
#> [545,]  79   1      0        180        1   170
#> [546,]  81  15      0        180        1   140
#> [547,]  80   7      1          0        1   146
#> [548,]  84   5      1          1        1    85
#> [549,]  83   8      0          8        0   115
#> [550,]  81  16      0         16        1   110
#> [551,]  80   6      1          0        1   150
#> [552,]  81   8      0        180        0   146
#> [553,]  79   7      0        177        0   197
#> [554,]  79   0      1          0        1    96
#> [555,]  85   4      0        180        0    90
#> [556,]  81   2      1          1        0   198
#> [557,]  83   2      0          2        1   155
#> [558,]  84   4      0        167        0   198
#> [559,]  80   3      1          1        1   230
#> [560,]  82  23      1          0        0   110
#> [561,]  84   5      0        180        1   203
#> [562,]  84   4      0          4        1    85
#> [563,]  81   1      0          1        1   150
#> [564,]  84   1      0         38        1   205
#> [565,]  83   3      0        180        0   174
#> [566,]  81   4      0         90        1   138
#> [567,]  79   9      1          8        0   150
#> [568,]  85   3      1          2        1   160
#> [569,]  80  13      1          8        1   140
#> [570,]  80   2      1          0        1   130
#> [571,]  80   6      0         71        1   189
#> [572,]  83   1      0          1        1   100
#> [573,]  80  30      1         13        0   220
#> [574,]  83   9      0        180        0   198
#> [575,]  79  14      1          0        0   110
#> [576,]  81  14      1         12        1   128
#> [577,]  82   0      0          2        1   100
#> [578,]  85   9      1          6        1   160
#> [579,]  81   4      0          4        0   175
#> [580,]  81   1      0          1        1   145
#> [581,]  82  16      1          8        0   103
#> [582,]  82   5      1          0        1   146
#> [583,]  83  12      1          2        1   170
#> [584,]  81  19      1         14        0   120
#> [585,]  80   2      0         88        0   135
#> [586,]  83   7      0        126        0   135
#> [587,]  86   8      0          8        1   132
#> [588,]  81  16      1          9        0   180
#> [589,]  84   6      0        165        0   145
#> [590,]  86   3      0          3        1   140
#> [591,]  82   9      0        180        1   134
#> [592,]  81  13      0        180        0   152
#> [593,]  85   3      0          3        1   118
#> [594,]  83   9      0        180        1   149
#> [595,]  83   4      0          4        0   130
#> [596,]  87   2      0          5        1   137
#> [597,]  86  12      1          0        1   132
#> [598,]  82  14      1         11        1   103
#> [599,]  86   6      1          0        1   140
#> [600,]  83  10      1          0        1   190
#> [601,]  86   2      0        180        1   169
#> [602,]  88  14      1          3        1   130
#> [603,]  84   3      0          3        1   121
#> [604,]  83  13      1         12        0   170
#> [605,]  87   2      0        180        0   113
#> [606,]  84   9      0         92        1   110
#> [607,]  84   3      0        180        1   170
#> [608,]  86   4      0         38        1   122
#> [609,]  82   4      0          4        0   130
#> [610,]  86  13      0        177        0   163
#> [611,]  85   3      0          3        1   113
#> [612,]  86   6      0          6        1   117
#> [613,]  84  13      0         62        1   100
#> [614,]  86   6      1          1        0   112
#> [615,]  83  20      1          3        1   150
#> [616,]  88   4      0          4        1   115
#> [617,]  85  22      0         22        1   184
#> [618,]  83   9      0         65        1   150
#> [619,]  87   2      0        180        1   130
#> [620,]  86   6      0         46        0   173
#> [621,]  88   3      0        115        0   110
#> [622,]  86  15      1          8        1   109
#> [623,]  89   4      0          4        1   153
#> [624,]  89   5      0        119        1   140
#> [625,]  87   6      0        180        1   110
#> [626,]  87   1      0          1        0   170
#> [627,]  84   8      0        180        1   119
#> [628,]  87  29      0         29        1    97
#> [629,]  87  15      1          9        1   138
#> [630,]  84   0      0        180        1   136
#> [631,]  89  10      0         46        1   170
#> [632,]  88   1      0          1        0   135
#> [633,]  86   4      0        180        1   145
#> [634,]  91   8      0          8        0   100
#> [635,]  87   2      0        180        0   160
#> [636,]  87   6      1          0        0   125
#> [637,]  91  10      0        145        0   135
#> [638,]  86   3      1          0        1    80
#> [639,]  88   7      0         24        0   119
#> [640,]  88   8      0         50        1   154
#> [641,]  86  10      0        180        1   137
#> [642,]  86   9      1          7        0   130
#> [643,]  90   4      1          0        0   121
#> [644,]  91   1      0          1        1    74
#> [645,]  87   5      0         36        1   150
#> [646,]  90   5      1          0        1   125
#> [647,]  89   3      1          1        1   160
#> [648,]  92   1      0          1        1   167
#> [649,]  87   7      0         74        1   105
#> [650,]  89  12      1          0        1   130
#> [651,]  91   5      0        169        1   176
#> [652,]  89  52      0         52        1   130
#> [653,]  92   7      0          7        1   110
#> [654,]  91   0      0          0        0     0
#> [655,]  91   4      1          0        1   120
#> [656,]  90  19      1         11        1   129
#> [657,]  94   6      0         50        0    78
#> [658,]  90   1      0          1        1   118
#> [659,]  91   2      0          2        1   116
#> [660,]  92   4      0         76        1   149
#> [661,]  90  16      0         16        1   106
#> [662,]  90   3      0         67        0   162
#> [663,]  96   3      0         12        1    97
#> [664,]  94   3      0         26        1   144
#> [665,]  91   7      0          7        0   135
#> [666,]  93   0      1          0        1   122
#> [667,]  92   2      0          2        0   112
#> [668,]  93   4      0        180        1   135
#> [669,]  96   3      1          0        1   104
#> [670,]  96  15      1          0        1   140
#> 
#> $y
#>   [1] 180.0+   2.0+ 180.0+   5.0+ 180.0+   2.0+ 180.0+ 180.0+ 115.0  180.0+
#>  [11]   5.0+ 180.0+ 180.0+ 180.0+   2.0+   5.0+   2.0+   3.0  180.0+ 180.0+
#>  [21] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [31] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [41] 180.0+ 180.0+ 180.0+  73.0    5.0+ 161.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [51]   5.0+ 180.0+ 180.0+ 180.0+ 177.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [61]  10.0+ 172.0+ 180.0+ 180.0+ 180.0+ 180.0+   1.0  180.0+ 179.0+ 180.0+
#>  [71] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  36.0 
#>  [81]  88.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0  180.0+ 180.0+
#>  [91] 180.0+ 180.0+ 180.0+  85.0  180.0+ 166.0+  99.0   16.0+ 180.0+   7.0+
#> [101]   6.0+ 180.0+ 180.0+  13.0+ 171.0+ 180.0+ 174.0+   6.0+   1.0  180.0+
#> [111] 180.0+ 180.0+ 180.0+ 175.0+ 180.0+   2.0  180.0+   7.0+ 180.0+ 180.0+
#> [121] 180.0+  16.0+ 180.0+  16.0   15.0+ 180.0+ 180.0+ 180.0+  12.0+   8.0 
#> [131]   2.0    3.0+ 180.0+ 180.0+ 180.0+ 140.0  180.0+ 165.0  180.0+ 180.0+
#> [141]   0.5  180.0+ 180.0+ 180.0+ 171.0+  15.0    5.0+   4.0+ 180.0+ 180.0+
#> [151]   5.0+ 180.0+   4.0+   9.0+   1.0    2.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [161]   2.0  180.0+ 180.0+ 180.0+  64.0  180.0+ 180.0+ 180.0+   0.5  180.0+
#> [171] 161.0+ 171.0+ 180.0+ 180.0+   3.0    1.0  180.0+ 180.0+ 180.0+   3.0+
#> [181]  19.0  180.0+   9.0+  24.0    8.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [191]  77.0   13.0+ 180.0+ 180.0+ 180.0+  94.0  180.0+   7.0+   6.0  180.0+
#> [201]   3.0+  30.0  180.0+ 180.0+  28.0    5.0   18.0    5.0+ 180.0+ 180.0+
#> [211]   1.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   9.0    7.0+
#> [221]  22.0   84.0    7.0+ 180.0+   1.0    1.0  180.0+   4.0+   6.0+ 180.0+
#> [231] 180.0+ 180.0+  12.0  180.0+  36.0  180.0+   3.0+  88.0  180.0+  12.0 
#> [241] 180.0+ 180.0+ 180.0+ 180.0+   0.5  180.0+  12.0+ 180.0+ 180.0+ 180.0+
#> [251] 180.0+ 180.0+  12.0    9.0  180.0+ 180.0+ 180.0+ 180.0+   3.0+ 103.0 
#> [261]  15.0  180.0+   5.0+  13.0  179.0+   0.5+ 180.0+   3.0+ 175.0+ 180.0+
#> [271]   7.0+   8.0    1.0  180.0+ 180.0+ 180.0+ 123.0+   1.0+  18.0  180.0+
#> [281]  79.0  180.0+   4.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   3.0  175.0 
#> [291]  10.0  180.0+ 180.0+ 180.0+   6.0   19.0+  12.0   99.0  180.0+ 180.0+
#> [301]   0.5    7.0+ 180.0+ 152.0+ 180.0+ 180.0+  93.0  180.0+   1.0   18.0+
#> [311] 101.0    4.0    5.0    7.0+ 150.0  180.0+ 180.0+ 180.0+ 180.0+   6.0 
#> [321] 180.0+   0.5  180.0+ 180.0+  14.0+ 180.0+   7.0+   2.0   45.0    5.0+
#> [331] 180.0+   5.0+ 180.0+   7.0    8.0+ 180.0+ 180.0+ 172.0+ 180.0+   8.0+
#> [341]  13.0+ 180.0+ 180.0+  19.0  180.0+   1.0    1.0   60.0  132.0   10.0+
#> [351] 180.0+ 162.0    7.0+   7.0+   9.0  180.0+ 180.0+  12.0  180.0+ 180.0+
#> [361] 180.0+   2.0   76.0  180.0+ 173.0+ 180.0+ 180.0+  28.0  180.0+  16.0+
#> [371] 180.0+  16.0+ 180.0+ 180.0+   7.0+  15.0    3.0+  13.0+ 180.0+   2.0 
#> [381]   3.0+  20.0  180.0+  20.0    0.5  180.0+ 180.0+ 180.0+   3.0   12.0 
#> [391]  58.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   3.0  175.0    2.0 
#> [401] 180.0+  14.0+  10.0+ 179.0+ 159.0   10.0    1.0   13.0  180.0+ 180.0+
#> [411]  10.0  104.0+   1.0  180.0+  11.0    5.0  180.0+  12.0  180.0+ 180.0+
#> [421] 180.0+ 180.0+ 180.0+  34.0  180.0+ 177.0+ 180.0+  11.0+   4.0+  15.0+
#> [431] 180.0+  10.0  180.0+ 180.0+ 180.0+ 180.0+   6.0  180.0+ 180.0+  85.0 
#> [441] 180.0+  17.0+   7.0    0.5  180.0+ 180.0+  46.0    1.0  180.0+ 180.0+
#> [451]  12.0  180.0+ 180.0+ 180.0+   8.0  180.0+   5.0  180.0+ 180.0+ 180.0+
#> [461] 180.0+   7.0+  79.0    3.0  168.0+ 180.0+ 180.0+ 176.0+ 180.0+  47.0 
#> [471] 180.0+  11.0    7.0  180.0+   8.0+ 180.0+  32.0  180.0+  10.0  180.0+
#> [481] 172.0  119.0   12.0  180.0+   8.0    1.0   80.0  180.0+ 180.0+   4.0+
#> [491] 152.0+   3.0    6.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0   10.0+
#> [501] 180.0+   6.0    3.0+   2.0+ 180.0+   1.0  171.0   43.0    6.0  138.0 
#> [511] 180.0+  71.0   59.0   17.0  161.0  180.0+ 164.0  173.0  180.0+  37.0 
#> [521] 180.0+   5.0+ 180.0+   3.0  171.0+  20.0+  10.0   10.0    6.0+   6.0 
#> [531]   5.0  180.0+ 180.0+ 108.0  125.0  180.0+ 180.0+ 180.0+ 180.0+ 103.0 
#> [541] 180.0+ 177.0+ 169.0   70.0  180.0+ 180.0+   7.0+ 180.0+   8.0+  16.0 
#> [551] 180.0+ 180.0+ 177.0+   0.5  180.0+ 180.0+   2.0  167.0    3.0+  62.0 
#> [561] 180.0+   4.0    1.0   38.0  180.0+  90.0  180.0+ 180.0+ 180.0+ 180.0+
#> [571]  71.0    1.0   30.0  180.0+ 180.0+ 180.0+   2.0  180.0+   4.0+   1.0 
#> [581]  16.0+   5.0+  77.0  180.0+  88.0  126.0    8.0  180.0+ 165.0    3.0 
#> [591] 180.0+ 180.0+   3.0+ 180.0+   4.0+   5.0  180.0+ 174.0    6.0  180.0+
#> [601] 180.0+  14.0    3.0   13.0  180.0+  92.0  180.0+  38.0    4.0  177.0 
#> [611]   3.0+   6.0+  62.0    6.0+  20.0    4.0   22.0   65.0  180.0+  46.0 
#> [621] 115.0  180.0+   4.0  119.0  180.0+   1.0+ 180.0+  29.0  180.0+ 180.0+
#> [631]  46.0    1.0+ 180.0+   8.0  180.0+  25.0  145.0    3.0   24.0   50.0 
#> [641] 180.0+ 180.0+   4.0    1.0   36.0   89.0    3.0+   1.0   74.0  180.0+
#> [651] 169.0   52.0    7.0    0.5    4.0  180.0+  50.0    1.0+   2.0   76.0 
#> [661]  16.0   67.0   12.0   26.0    7.0+   0.5    2.0  180.0+   3.0   15.0+
#> 
#> $weights
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8473076