Cross-Validated GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::cv.glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.cv.glmnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.cv.glmnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.cv_glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

foldid	untyped	NULL		-

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nfolds	integer	10		\([3, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	TRUE	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	lambda.1se		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

standardize.response	logical	FALSE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.gaussian	character	-	covariance, naive	-

type.logistic	character	Newton	Newton, modified.Newton	-

type.measure	character	deviance	deviance, C	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.coxnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvCVGlmnet

Methods

Public methods

LearnerSurvCVGlmnet$new()

LearnerSurvCVGlmnet$selected_features()

LearnerSurvCVGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvCVGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvCVGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvCVGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.cv_glmnet")
print(learner)
#> 
#> ── <LearnerSurvCVGlmnet> (surv.cv_glmnet): Regularized Generalized Linear Model 
#> • Model: -
#> • Parameters: use_pred_offset=TRUE
#> • Packages: mlr3, mlr3proba, mlr3extralearners, and glmnet
#> • Predict Types: [crank], distr, and lp
#> • Feature Types: logical, integer, and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: offset, selected_features, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#> Measure: Partial Likelihood Deviance 
#> 
#>      Lambda Index Measure     SE Nonzero
#> min 0.00313    45   2.764 0.1822       6
#> 1se 0.07392    11   2.919 0.1989       3
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  32   5      1          0        1   121
#>   [2,]  33   2      0          2        0   150
#>   [3,]  35  10      1          9        0   106
#>   [4,]  35   2      0        180        0   121
#>   [5,]  37   9      0        180        1   151
#>   [6,]  38  13      1          0        1   161
#>   [7,]  35   0      0        180        1   119
#>   [8,]  38  12      1          8        1   120
#>   [9,]  38  16      1         10        0   160
#>  [10,]  38  12      1         11        1    92
#>  [11,]  40  12      1          9        0   153
#>  [12,]  42   3      1          1        1   130
#>  [13,]  40   2      1          1        1   148
#>  [14,]  38   5      1          3        0   125
#>  [15,]  42   2      0          2        0   140
#>  [16,]  40   6      0        180        1   138
#>  [17,]  40  11      1         10        1   120
#>  [18,]  42   2      0        180        0   100
#>  [19,]  40   1      1          0        1   145
#>  [20,]  42   4      0        180        0   162
#>  [21,]  42  12      1         10        1   170
#>  [22,]  43   2      1          1        1   116
#>  [23,]  42   2      0        180        1   124
#>  [24,]  44   5      1          1        0   170
#>  [25,]  41  10      1          8        0   150
#>  [26,]  44   3      0        180        0   141
#>  [27,]  41  13      1          1        0   140
#>  [28,]  45   9      1          7        0   110
#>  [29,]  45   6      0        180        1   170
#>  [30,]  41   5      1          4        1   141
#>  [31,]  44   2      1          1        1   150
#>  [32,]  43   2      0        180        1   140
#>  [33,]  45   2      0        180        1   140
#>  [34,]  46  15      0        180        0   120
#>  [35,]  46   2      1          1        0   126
#>  [36,]  47   4      1          3        0   118
#>  [37,]  48  15      0        180        1   160
#>  [38,]  45   3      0        150        0   130
#>  [39,]  44   3      1          0        1   180
#>  [40,]  46   7      1          2        0   166
#>  [41,]  43  29      0        180        1   180
#>  [42,]  47   6      1          0        1   116
#>  [43,]  44   0      1          0        1    96
#>  [44,]  47   4      1          3        1   160
#>  [45,]  43   3      1          0        1   124
#>  [46,]  49   5      0         73        1   136
#>  [47,]  45   5      0          5        0   141
#>  [48,]  46   6      1          0        1   100
#>  [49,]  47   2      0        180        0   108
#>  [50,]  45   5      0        180        1   190
#>  [51,]  46   5      1          3        0   130
#>  [52,]  46   4      0        180        1   121
#>  [53,]  44   2      0        180        0   142
#>  [54,]  46  15      0        180        1   120
#>  [55,]  47   3      1          1        1   120
#>  [56,]  48   3      0        180        0   154
#>  [57,]  48  12      1         11        0   200
#>  [58,]  47   5      1          3        1   130
#>  [59,]  47   9      1          6        0   170
#>  [60,]  49   4      0        180        0   117
#>  [61,]  47  10      0         10        1   140
#>  [62,]  50   1      1          0        1   129
#>  [63,]  48   2      1          0        0   184
#>  [64,]  50   4      1          1        0   125
#>  [65,]  49   7      1          7        1   110
#>  [66,]  46   3      1          1        1   140
#>  [67,]  50   7      0        180        1   110
#>  [68,]  49  15      1         11        1   160
#>  [69,]  47   2      0        180        0   150
#>  [70,]  46   6      1          0        1   156
#>  [71,]  50   7      1          0        1    92
#>  [72,]  51   3      1          2        0   113
#>  [73,]  50   1      1          0        0   150
#>  [74,]  50   9      0        180        0   130
#>  [75,]  47   8      0        180        0   160
#>  [76,]  47   6      0        180        1   162
#>  [77,]  51   8      0        180        1   140
#>  [78,]  52   2      0        180        0   155
#>  [79,]  46   3      0        180        1   120
#>  [80,]  46   1      1          1        0   145
#>  [81,]  50   4      1          1        0   150
#>  [82,]  48   7      1          0        1   110
#>  [83,]  53   8      0         36        1   160
#>  [84,]  48  17      1         10        0   111
#>  [85,]  47   2      1          1        0   110
#>  [86,]  52   4      1          4        0   152
#>  [87,]  49   9      1          3        0   102
#>  [88,]  49  15      0        180        1   160
#>  [89,]  54  17      1         12        1   102
#>  [90,]  53   5      0         77        0   159
#>  [91,]  53   7      1          0        0   199
#>  [92,]  51   3      1          1        0   140
#>  [93,]  50   2      0          5        1   106
#>  [94,]  50  14      1         13        0   170
#>  [95,]  53   8      1          7        0   160
#>  [96,]  51  25      1          1        0   202
#>  [97,]  53   4      0          4        0   140
#>  [98,]  52  14      1          7        1   200
#>  [99,]  48   6      0        180        0   160
#> [100,]  53   4      1          0        1   156
#> [101,]  51  13      0         99        1   160
#> [102,]  49  16      0         16        0   125
#> [103,]  55   3      1          1        0   150
#> [104,]  54  23      1         10        0   131
#> [105,]  52   7      1          2        0   154
#> [106,]  54   9      1          1        0   130
#> [107,]  55   4      1          2        0   150
#> [108,]  51  13      1         11        0   145
#> [109,]  50   5      1          4        1   150
#> [110,]  50   3      0        174        1   153
#> [111,]  55  28      1         13        1   160
#> [112,]  50   7      1          1        0   156
#> [113,]  53   9      0          9        1    95
#> [114,]  53   8      1          0        1   130
#> [115,]  52   5      0        175        1   117
#> [116,]  55   1      0        180        0   127
#> [117,]  55   2      0          2        0   145
#> [118,]  56   2      0        180        0   132
#> [119,]  55   5      1          4        1   120
#> [120,]  54   3      0        180        1   180
#> [121,]  55   6      0        180        0   170
#> [122,]  52  16      0         16        0   152
#> [123,]  55   6      0        180        1   100
#> [124,]  55   6      1          5        1   138
#> [125,]  54  12      1          0        1   190
#> [126,]  55   2      0        134        1   140
#> [127,]  56   3      0          8        1   139
#> [128,]  55   1      0          2        0   130
#> [129,]  54   7      1          2        0   129
#> [130,]  54   2      1          1        0   135
#> [131,]  52   9      1          3        0   170
#> [132,]  54   2      1          1        1   176
#> [133,]  57   5      1          3        1   138
#> [134,]  57   1      0        180        1   156
#> [135,]  57   1      0          1        1   100
#> [136,]  56   4      1          0        1   140
#> [137,]  52   2      0        180        0   140
#> [138,]  55  11      1          7        0   104
#> [139,]  52  15      1         14        0   130
#> [140,]  56  14      1         11        0   130
#> [141,]  58   8      0          8        1   130
#> [142,]  54   5      0        180        1   108
#> [143,]  55   3      1          1        1   156
#> [144,]  57   0      0          0        1   150
#> [145,]  53  21      1         13        1   130
#> [146,]  59   3      1          1        0   172
#> [147,]  57   4      0        180        1   119
#> [148,]  53  15      1         10        1   130
#> [149,]  55  13      0        166        1   140
#> [150,]  56   5      0          5        1   150
#> [151,]  54  23      1          8        0   120
#> [152,]  57   4      1          2        1   185
#> [153,]  53   4      0        147        1   145
#> [154,]  57  11      1         10        1   129
#> [155,]  55   3      1          2        0   140
#> [156,]  55   5      0          5        1   131
#> [157,]  54   7      1          0        1   141
#> [158,]  56   4      0          4        0   164
#> [159,]  59  15      1         10        0   140
#> [160,]  58   9      1          0        1   180
#> [161,]  58   1      1          1        1   200
#> [162,]  59   9      1          1        1   125
#> [163,]  57   1      0        180        0   148
#> [164,]  60  11      1          9        0   106
#> [165,]  59   3      0        180        0   120
#> [166,]  58   4      1          0        1   160
#> [167,]  57   2      0          2        1   120
#> [168,]  60   5      1          1        0   138
#> [169,]  58  11      1          9        1   124
#> [170,]  55   5      1          0        1   160
#> [171,]  57  10      1          9        0   103
#> [172,]  59   5      0        180        1   155
#> [173,]  59   4      1          0        1   152
#> [174,]  58  26      1          0        1   189
#> [175,]  61   9      0          9        1   160
#> [176,]  58   4      1          3        0   120
#> [177,]  60   0      1          0        1    80
#> [178,]  59   2      1          1        0   140
#> [179,]  58  14      1          6        0   190
#> [180,]  61   4      1          3        0   151
#> [181,]  61   9      1          8        0   150
#> [182,]  58   1      0          1        1   100
#> [183,]  61  20      1         13        0   130
#> [184,]  57   2      1          0        1   116
#> [185,]  58  10      0         10        1   150
#> [186,]  57   4      1          3        0   138
#> [187,]  61   3      0         17        0   143
#> [188,]  56  14      0         45        0   130
#> [189,]  57   3      1          2        0   120
#> [190,]  55   4      1          3        1   160
#> [191,]  58  11      0        172        1   135
#> [192,]  60  12      1          0        1   114
#> [193,]  61   4      1          0        1   115
#> [194,]  56   8      1          8        0   120
#> [195,]  61  13      1         12        1   130
#> [196,]  59  11      1          8        1   190
#> [197,]  57  15      1         13        1   110
#> [198,]  59   5      1          2        0   182
#> [199,]  59  10      0        180        0   160
#> [200,]  61   8      0         77        0   120
#> [201,]  61  13      0         13        0   210
#> [202,]  58   8      1          5        0   152
#> [203,]  62  10      1          0        1   153
#> [204,]  57   3      1          0        0   100
#> [205,]  61  18      0        170        0   140
#> [206,]  61  28      1          7        0   133
#> [207,]  58   8      1          3        1   150
#> [208,]  57   7      0        169        0   180
#> [209,]  61   7      0          7        1   150
#> [210,]  60   7      0          7        0   147
#> [211,]  59  13      1          2        0   198
#> [212,]  60  17      1          8        1   140
#> [213,]  58   3      1          0        1   146
#> [214,]  58   2      0         30        0   202
#> [215,]  61   5      0          5        1   110
#> [216,]  61   5      0          5        1   160
#> [217,]  58  11      1          9        0   179
#> [218,]  62  17      1         10        1   180
#> [219,]  63   3      1          1        0   180
#> [220,]  62   3      0        180        1   105
#> [221,]  63   4      0        180        1   190
#> [222,]  64   4      0        180        0   130
#> [223,]  63   4      1          1        0   155
#> [224,]  59   8      0        180        1   140
#> [225,]  58   9      1          9        0   110
#> [226,]  62   7      0          7        0   150
#> [227,]  59   1      0         22        1   162
#> [228,]  59   4      0        180        0   196
#> [229,]  60   7      1          5        1   141
#> [230,]  60   7      0          7        0   140
#> [231,]  59   5      1          1        0   148
#> [232,]  63   1      0          1        0   162
#> [233,]  63   1      0          1        0   130
#> [234,]  62   6      0        180        0   170
#> [235,]  61  15      1         13        0   170
#> [236,]  59   4      0          4        0   149
#> [237,]  60   3      0          3        0   168
#> [238,]  64  10      1          9        0   160
#> [239,]  63  12      1         10        0   200
#> [240,]  60   8      0         17        1   130
#> [241,]  61   6      1          1        1   117
#> [242,]  64  12      1         11        0   160
#> [243,]  64   6      1          0        1   140
#> [244,]  63  14      1          9        0   123
#> [245,]  65  36      1         11        0   140
#> [246,]  63   4      1          3        0   162
#> [247,]  61  10      1          2        1   194
#> [248,]  64  32      1          9        1   160
#> [249,]  63  12      1          9        0   114
#> [250,]  63   7      0        180        0   120
#> [251,]  65   8      1          0        0   168
#> [252,]  65  10      1          8        1   120
#> [253,]  60   6      0        180        0   130
#> [254,]  64  21      1         10        0   190
#> [255,]  61  12      1         11        0   154
#> [256,]  61   4      0        180        1   113
#> [257,]  65   3      0        180        1   190
#> [258,]  66   6      1          1        1   130
#> [259,]  63  12      0         12        1   150
#> [260,]  62   3      1          1        1   199
#> [261,]  65   6      0          9        0   112
#> [262,]  65   3      1          0        1    80
#> [263,]  63   5      1          4        0   170
#> [264,]  63   2      1          1        0   180
#> [265,]  62  13      1         11        0   180
#> [266,]  67  11      0         11        1   100
#> [267,]  64   2      0          2        0   201
#> [268,]  66  18      1          5        0   142
#> [269,]  61  15      1         10        0   130
#> [270,]  63   9      1          8        1   160
#> [271,]  63   3      1          2        0   120
#> [272,]  63   2      1          0        0   140
#> [273,]  64  19      1          8        1   160
#> [274,]  65   8      1          0        1   140
#> [275,]  67   6      0        180        1   170
#> [276,]  65  15      1         11        1   160
#> [277,]  64   6      1          0        1   125
#> [278,]  66  13      1          0        0   118
#> [279,]  65   3      0          3        0   105
#> [280,]  66   3      1          0        1   135
#> [281,]  65   2      1          1        1   170
#> [282,]  68   1      0        180        1   166
#> [283,]  64  10      1          9        1   110
#> [284,]  63  10      0         16        1   160
#> [285,]  66  14      0        180        0   130
#> [286,]  64   1      0          1        1   120
#> [287,]  68  18      0        180        1   260
#> [288,]  63   8      1          1        1   162
#> [289,]  63   1      1          0        1   155
#> [290,]  63  10      0         18        1   130
#> [291,]  66  12      1         10        1   150
#> [292,]  65  15      1         12        1   150
#> [293,]  66  11      1          0        0   100
#> [294,]  65   4      1          2        1   145
#> [295,]  69  12      0         15        1   140
#> [296,]  66  15      1         13        1   160
#> [297,]  63   2      0        180        0   150
#> [298,]  65  11      1          6        0   130
#> [299,]  69   6      0        180        1   100
#> [300,]  68  14      1         13        1   140
#> [301,]  65   8      1          0        1    90
#> [302,]  66   3      0          3        1   138
#> [303,]  67   1      0        180        1   160
#> [304,]  68  10      1         10        1   150
#> [305,]  65   1      1          0        0   133
#> [306,]  67   7      1          4        1   130
#> [307,]  63   2      1          0        0    99
#> [308,]  67   2      0        180        0   184
#> [309,]  65  10      1          1        1   148
#> [310,]  66  19      1         12        1   150
#> [311,]  69   6      0         99        1   140
#> [312,]  65   4      1          1        0   130
#> [313,]  64   4      0        179        0   160
#> [314,]  66   4      0        180        1   130
#> [315,]  64  11      0         11        0   125
#> [316,]  64   4      0        180        1   140
#> [317,]  64   0      1          0        1   118
#> [318,]  66   7      1          5        1   131
#> [319,]  66   4      0        180        0   177
#> [320,]  69   4      1          3        1   150
#> [321,]  65  13      1         12        1   130
#> [322,]  64  21      0         21        1   155
#> [323,]  66   6      0        180        0   140
#> [324,]  68  18      1          0        1   160
#> [325,]  65   6      0        101        1   115
#> [326,]  68   4      0          4        1   190
#> [327,]  68   7      0        150        0   210
#> [328,]  71  20      1          0        1   160
#> [329,]  67   2      0        180        0   128
#> [330,]  66   9      1          3        1   151
#> [331,]  66   1      1          1        1   165
#> [332,]  70   4      1          0        1   180
#> [333,]  69   8      0        180        1   153
#> [334,]  70  14      0        171        0   166
#> [335,]  66   4      0        180        0   130
#> [336,]  67  10      1          9        0   200
#> [337,]  65   2      0        180        0   130
#> [338,]  69   3      1          2        0   151
#> [339,]  65  14      1         13        1   150
#> [340,]  71   7      0          7        0   230
#> [341,]  66   2      0          2        1   228
#> [342,]  70  22      1         13        0   103
#> [343,]  67   1      0         36        1   104
#> [344,]  72   3      1          0        1   132
#> [345,]  72   7      0          7        1   110
#> [346,]  69   8      1          7        1   108
#> [347,]  67   3      0        180        0   110
#> [348,]  66   2      1          1        0   123
#> [349,]  69  19      0        180        0   130
#> [350,]  68  18      0         18        1   100
#> [351,]  69  11      1          0        1   120
#> [352,]  66   2      0        180        0   130
#> [353,]  69   4      1          3        0   132
#> [354,]  68   2      0          7        1   130
#> [355,]  69   8      1          2        0   121
#> [356,]  70   3      0        123        0   130
#> [357,]  72   5      1          4        0   170
#> [358,]  67  22      1          1        1   140
#> [359,]  68   3      0         19        0   135
#> [360,]  69   1      0          1        1   110
#> [361,]  69   5      0         76        0   120
#> [362,]  67   8      1          0        1   130
#> [363,]  66  24      1         13        0   130
#> [364,]  72  30      1          0        1   145
#> [365,]  70   7      0          7        0   102
#> [366,]  73  20      1          0        1   170
#> [367,]  71   6      0          9        0   120
#> [368,]  70  11      0        180        1   210
#> [369,]  72  19      1          8        0   120
#> [370,]  72  12      1         10        0   170
#> [371,]  67   8      0        180        1   170
#> [372,]  67   9      0        180        0   158
#> [373,]  72   2      0          2        1   100
#> [374,]  67   4      1          1        0   134
#> [375,]  72   6      1          5        0   115
#> [376,]  71   1      0        173        1   188
#> [377,]  70   3      0        180        0   121
#> [378,]  71   3      1          2        0   150
#> [379,]  68   4      1          3        0   210
#> [380,]  72   5      0         28        0   120
#> [381,]  71   5      0        180        0   191
#> [382,]  73   6      0        180        1   117
#> [383,]  69  16      1         10        1   140
#> [384,]  69   8      1          1        0   164
#> [385,]  68   7      0        180        1   130
#> [386,]  72  16      1          1        1   130
#> [387,]  70   4      0        180        0   180
#> [388,]  73   6      1          0        1   270
#> [389,]  72   8      1          1        1   150
#> [390,]  71   2      1          0        1   180
#> [391,]  73   7      0          7        1   140
#> [392,]  68  15      1         13        1   130
#> [393,]  70   3      0          3        1   159
#> [394,]  70  13      1          9        0   100
#> [395,]  72   6      0        180        1   130
#> [396,]  69   2      1          0        1   110
#> [397,]  71   3      1          1        0   150
#> [398,]  71  15      1         11        0   165
#> [399,]  74  20      0         20        1   180
#> [400,]  68   9      0        180        1   120
#> [401,]  73   3      1          0        1   136
#> [402,]  70   5      1          0        1   190
#> [403,]  71  17      1         11        0   160
#> [404,]  71   3      1          2        1   190
#> [405,]  73  10      1          8        0   106
#> [406,]  70  26      1         11        1   120
#> [407,]  74   4      0          4        0   120
#> [408,]  72   5      1          3        1   160
#> [409,]  70   3      0        180        1   154
#> [410,]  72  15      1          0        1   150
#> [411,]  71   7      1          2        0   143
#> [412,]  72   8      1          0        1   140
#> [413,]  74   3      0          3        1   150
#> [414,]  73  17      1         11        0   140
#> [415,]  71  13      1          8        0   121
#> [416,]  69   2      1          1        1    80
#> [417,]  71  14      1         13        1   170
#> [418,]  74   7      1          0        1   117
#> [419,]  69   7      0        180        1   144
#> [420,]  72  15      1         13        0   156
#> [421,]  71  10      1          9        1   120
#> [422,]  75   2      1          1        0   145
#> [423,]  73  10      1          9        1   146
#> [424,]  72  10      1          9        1   160
#> [425,]  74  15      1          9        1   179
#> [426,]  71   2      0         10        1   112
#> [427,]  73   1      0          1        1    80
#> [428,]  73  10      1          8        0   120
#> [429,]  72   7      0         57        1   145
#> [430,]  73  10      0        180        0   162
#> [431,]  72  11      0         11        1   140
#> [432,]  70   3      0          3        0   150
#> [433,]  73   5      1          3        1   112
#> [434,]  76  25      1         12        1   170
#> [435,]  73  12      1         12        1   140
#> [436,]  72   2      0        180        0   120
#> [437,]  75   1      0        180        1   140
#> [438,]  72   4      1          0        1   197
#> [439,]  71   3      1          0        0   144
#> [440,]  73   5      0        180        0   126
#> [441,]  74  34      1          8        1   233
#> [442,]  71  32      1         12        1   107
#> [443,]  72   5      0        180        0   154
#> [444,]  72   3      0        180        0   160
#> [445,]  77  11      0         11        1   150
#> [446,]  77   4      0          4        0   185
#> [447,]  72   7      1          2        0   142
#> [448,]  73  15      0         15        1   160
#> [449,]  71  16      0        180        0   140
#> [450,]  73  10      1         10        0   124
#> [451,]  74   7      0        180        1   150
#> [452,]  76   1      0        180        0   114
#> [453,]  74   2      1          1        0   140
#> [454,]  76   8      1          0        1   141
#> [455,]  74  19      1          4        1   200
#> [456,]  73   6      0          6        1   114
#> [457,]  75  23      1         14        1   110
#> [458,]  74   2      0        180        0   190
#> [459,]  72   4      0         85        1   120
#> [460,]  72   4      1          3        0   160
#> [461,]  76  17      1          0        1   200
#> [462,]  73   4      1          3        1   125
#> [463,]  76  13      1         10        0   110
#> [464,]  75   7      0          7        0   190
#> [465,]  75   0      0          0        1   130
#> [466,]  75  12      0         12        1   160
#> [467,]  76  13      1          8        1   148
#> [468,]  75   4      1          2        1   188
#> [469,]  74   6      0        180        0   160
#> [470,]  76   4      0          4        1   155
#> [471,]  75   1      0          1        1   125
#> [472,]  74   2      0        180        0   111
#> [473,]  73   1      0         52        1   105
#> [474,]  73   0      0        180        0   156
#> [475,]  72   5      0        180        0   120
#> [476,]  76  44      1         10        0   105
#> [477,]  74  10      1          0        1   135
#> [478,]  77   5      1          0        0   123
#> [479,]  77  12      0        180        0   130
#> [480,]  77   1      1          0        1    90
#> [481,]  76  12      1         11        1   120
#> [482,]  78   5      1          0        1   170
#> [483,]  74   6      0         79        1   140
#> [484,]  75   3      1          1        1   171
#> [485,]  74   9      1          8        0   200
#> [486,]  74   2      1          0        1   130
#> [487,]  77   3      0        180        0   110
#> [488,]  76  29      0         47        0    90
#> [489,]  73  11      1          2        1   110
#> [490,]  78   7      0          7        1   133
#> [491,]  76  13      1          1        1   170
#> [492,]  78  32      1          9        1   198
#> [493,]  79   6      0        180        0   170
#> [494,]  78  13      1          5        0   130
#> [495,]  75   5      0        119        1   150
#> [496,]  75  12      1          1        1   120
#> [497,]  78  15      0        180        1   270
#> [498,]  75  13      1          6        0   150
#> [499,]  76   1      0          1        1    83
#> [500,]  79   4      0         80        0   145
#> [501,]  78  12      1          9        0   150
#> [502,]  77   2      1          0        1   143
#> [503,]  78  10      0        180        1   130
#> [504,]  76  11      1          0        0   120
#> [505,]  75   3      0          3        0     0
#> [506,]  77  24      0         24        1   160
#> [507,]  79   8      0         32        1   120
#> [508,]  80   9      0         23        1   128
#> [509,]  80   6      0          6        1   150
#> [510,]  78   6      1          0        1   240
#> [511,]  76   3      1          0        1   140
#> [512,]  79  11      0        180        0   160
#> [513,]  81   1      0          1        0   130
#> [514,]  78  11      1          8        1   118
#> [515,]  76  10      1          8        0   180
#> [516,]  80   3      1          0        1   120
#> [517,]  75   2      1          1        1   204
#> [518,]  77   7      0        180        1   170
#> [519,]  79   4      1          0        1   120
#> [520,]  81   1      0        180        0   120
#> [521,]  77   9      1          4        0   141
#> [522,]  82   5      0          8        1   120
#> [523,]  79  10      0         10        1   120
#> [524,]  80  15      1          0        1    90
#> [525,]  79  28      0        164        0   100
#> [526,]  80   9      0        118        1   186
#> [527,]  80   6      0        173        1   160
#> [528,]  78  32      0        180        1   130
#> [529,]  79   1      0         37        1   140
#> [530,]  81   3      0        180        0   184
#> [531,]  78   7      0          7        1   147
#> [532,]  77  13      1          0        1   190
#> [533,]  80   5      1          1        1   108
#> [534,]  78   4      0        180        0   175
#> [535,]  79   3      0          3        1   101
#> [536,]  76   1      0        166        0   131
#> [537,]  81   4      1          1        1   104
#> [538,]  78  20      1          0        1   109
#> [539,]  80   1      0          1        0   100
#> [540,]  82   3      1          1        1   144
#> [541,]  77   5      0         85        0   188
#> [542,]  79   6      0          6        0   152
#> [543,]  80   6      1          0        1   119
#> [544,]  78   2      0        180        0   148
#> [545,]  80   5      0          5        1   130
#> [546,]  82   1      1          0        1    82
#> [547,]  79  10      0        180        1   150
#> [548,]  77   4      0        180        1    98
#> [549,]  81   1      0        108        0   129
#> [550,]  78  12      0        180        0   134
#> [551,]  82  21      1          2        0   155
#> [552,]  79   4      0          4        1   121
#> [553,]  80   6      0          6        1   110
#> [554,]  82   5      0        180        0   110
#> [555,]  79   7      1          6        0   130
#> [556,]  83   4      0        103        0    97
#> [557,]  81   5      0        177        0    41
#> [558,]  80  11      1          8        0   170
#> [559,]  78  23      1         10        1   145
#> [560,]  79   4      0          4        1   183
#> [561,]  79   1      0        180        1   170
#> [562,]  81  15      0        180        1   140
#> [563,]  80   7      1          0        1   146
#> [564,]  84   5      1          1        1    85
#> [565,]  81  20      1          9        0   170
#> [566,]  83   8      0          8        0   115
#> [567,]  81  16      0         16        1   110
#> [568,]  80   6      1          0        1   150
#> [569,]  80  11      1          8        0   110
#> [570,]  81   8      0        180        0   146
#> [571,]  79   7      0        177        0   197
#> [572,]  79   0      1          0        1    96
#> [573,]  83   2      0          2        1   155
#> [574,]  82  23      1          0        0   110
#> [575,]  84   5      0        180        1   203
#> [576,]  84   4      0          4        1    85
#> [577,]  83   3      0        180        0   174
#> [578,]  81   4      0         90        1   138
#> [579,]  79   9      1          8        0   150
#> [580,]  85   3      1          2        1   160
#> [581,]  80  13      1          8        1   140
#> [582,]  80   2      1          0        1   130
#> [583,]  79   4      0          4        1    60
#> [584,]  83   1      0          1        1   100
#> [585,]  80  30      1         13        0   220
#> [586,]  79  14      1          0        0   110
#> [587,]  83   3      0        114        0    98
#> [588,]  81  14      1         12        1   128
#> [589,]  83   2      0        154        0   130
#> [590,]  82   0      0          2        1   100
#> [591,]  85   9      1          6        1   160
#> [592,]  81   4      0          4        0   175
#> [593,]  84  15      1         13        1   110
#> [594,]  81   1      0          1        1   145
#> [595,]  81  12      0         12        1   163
#> [596,]  82  16      1          8        0   103
#> [597,]  86  12      0        180        1   120
#> [598,]  81  19      1         14        0   120
#> [599,]  80   2      0         88        0   135
#> [600,]  83   7      0        126        0   135
#> [601,]  81  16      1          9        0   180
#> [602,]  84   6      0        165        0   145
#> [603,]  84   3      0        180        1   120
#> [604,]  81  13      0        180        0   152
#> [605,]  85   3      0          3        1   118
#> [606,]  81   2      1          0        1   118
#> [607,]  82   1      0        180        1   193
#> [608,]  83   4      0          4        0   130
#> [609,]  87   2      0          5        1   137
#> [610,]  86  12      1          0        1   132
#> [611,]  84   3      1          2        0   125
#> [612,]  83  10      1          0        1   190
#> [613,]  88  14      1          3        1   130
#> [614,]  84   3      0          3        1   121
#> [615,]  84   7      1          2        0   148
#> [616,]  87   2      0        180        0   113
#> [617,]  84   3      0        180        1   170
#> [618,]  86   4      0         38        1   122
#> [619,]  82   4      0          4        0   130
#> [620,]  86  13      0        177        0   163
#> [621,]  88   4      0          4        0   100
#> [622,]  83  20      1          3        1   150
#> [623,]  88   4      0          4        1   115
#> [624,]  85  22      0         22        1   184
#> [625,]  83   9      0         65        1   150
#> [626,]  87   2      0        180        1   130
#> [627,]  88   3      0        115        0   110
#> [628,]  83   3      0          3        1   130
#> [629,]  87   8      0          8        1   157
#> [630,]  86  15      1          8        1   109
#> [631,]  89   4      0          4        1   153
#> [632,]  89   5      0        119        1   140
#> [633,]  84   2      0        110        1   174
#> [634,]  87  15      1          9        1   138
#> [635,]  84   0      0        180        1   136
#> [636,]  89  10      0         46        1   170
#> [637,]  90  14      0         14        1   100
#> [638,]  86   4      0        180        1   145
#> [639,]  91   8      0          8        0   100
#> [640,]  87   6      1          0        0   125
#> [641,]  88   7      0         24        0   119
#> [642,]  90  11      1         10        1   186
#> [643,]  87   6      0        126        1   168
#> [644,]  86  10      0        180        1   137
#> [645,]  90   4      1          0        0   121
#> [646,]  91   1      0          1        1    74
#> [647,]  87  43      0        178        1   130
#> [648,]  90   5      1          0        1   125
#> [649,]  88   3      1          2        0   159
#> [650,]  89   3      1          1        1   160
#> [651,]  91   3      0         33        1   137
#> [652,]  88   5      0        158        0   100
#> [653,]  87   7      0         74        1   105
#> [654,]  89  12      1          0        1   130
#> [655,]  89   2      0        168        0   118
#> [656,]  89  52      0         52        1   130
#> [657,]  89   4      0          4        1   159
#> [658,]  91   4      1          0        1   120
#> [659,]  94   6      0         50        0    78
#> [660,]  90   1      0          1        1   118
#> [661,]  91   2      0          2        1   116
#> [662,]  92   4      0         76        1   149
#> [663,]  91   1      0        180        0   158
#> [664,]  95   8      1          5        1   150
#> [665,]  94   3      0         26        1   144
#> [666,]  91  12      0         53        1   212
#> [667,]  91   7      0          7        0   135
#> [668,]  93   0      1          0        1   122
#> [669,]  92   5      0         69        0   139
#> [670,]  93   4      0        180        1   135
#> 
#> $y
#>   [1]   5.0+   2.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  12.0  180.0+ 180.0+
#>  [11] 180.0+ 180.0+   2.0+   5.0+   2.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [21] 180.0+   2.0+ 180.0+ 155.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0+
#>  [31] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 150.0  180.0+ 180.0+
#>  [41] 180.0+   6.0+ 180.0+ 180.0+ 180.0+  73.0    5.0+ 180.0+ 180.0+ 180.0+
#>  [51]   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [61]  10.0+ 172.0+ 180.0+ 180.0+   7.0  180.0+ 180.0+ 179.0+ 180.0+ 180.0+
#>  [71] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [81] 180.0+   7.0   36.0   88.0+ 180.0+   4.0+ 180.0+ 180.0+ 180.0+  77.0 
#>  [91] 180.0+ 180.0+   5.0  180.0+ 180.0+ 180.0+   4.0+  85.0  180.0+ 166.0+
#> [101]  99.0   16.0+ 180.0+ 152.0+   7.0+ 180.0+ 180.0+  13.0+ 171.0+ 174.0+
#> [111]  28.0  180.0+   9.0+ 180.0+ 175.0+ 180.0+   2.0  180.0+ 180.0+ 180.0+
#> [121] 180.0+  16.0+ 180.0+ 180.0+  12.0+ 134.0+   8.0    2.0  180.0+ 180.0+
#> [131] 180.0+ 180.0+ 140.0  180.0+   1.0  165.0  180.0+ 180.0+ 180.0+ 180.0+
#> [141]   8.0+ 180.0+ 180.0+   0.5  180.0+ 180.0+ 180.0+ 180.0+ 166.0+   5.0+
#> [151] 180.0+   4.0+ 147.0+ 180.0+ 180.0+   5.0+ 180.0+   4.0+ 180.0+   9.0+
#> [161]   1.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0  180.0+ 180.0+ 180.0+
#> [171] 180.0+ 180.0+ 180.0+ 180.0+   9.0+ 180.0+   0.5  180.0+ 171.0+ 180.0+
#> [181] 180.0+   1.0  180.0+ 180.0+  10.0+ 180.0+  17.0   45.0    3.0+ 180.0+
#> [191] 172.0+ 172.0+ 180.0+   8.0  180.0+ 180.0+  15.0  180.0+ 180.0+  77.0 
#> [201]  13.0+   8.0+ 180.0+ 180.0+ 170.0   94.0  180.0+ 169.0    7.0    7.0+
#> [211] 180.0+ 180.0+   3.0+  30.0    5.0    5.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [221] 180.0+ 180.0+   4.0+ 180.0+   9.0    7.0+  22.0  180.0+  84.0    7.0+
#> [231] 180.0+   1.0    1.0  180.0+ 180.0+   4.0+   3.0+ 167.0  180.0+  17.0 
#> [241] 180.0+  12.0  180.0+  14.0+  36.0  180.0+  88.0  180.0+  12.0  180.0+
#> [251] 180.0+ 180.0+ 180.0+ 180.0+  12.0+ 180.0+ 180.0+ 180.0+  12.0  180.0+
#> [261]   9.0    3.0  180.0+ 180.0+ 180.0+  11.0+   2.0+  18.0+ 180.0+ 180.0+
#> [271]   3.0+   2.0+ 103.0   15.0  180.0+ 180.0+ 180.0+ 166.0+   3.0    3.0+
#> [281] 175.0+ 180.0+ 180.0+  16.0  180.0+   1.0  180.0+ 180.0+   1.0+  18.0 
#> [291]  80.0   15.0+ 180.0+   4.0+  15.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [301]   8.0+   3.0  180.0+  10.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+  19.0+
#> [311]  99.0  180.0+ 179.0+ 180.0+  11.0+ 180.0+   0.5    7.0+ 180.0+ 152.0+
#> [321] 180.0+  21.0+ 180.0+  18.0+ 101.0    4.0  150.0  180.0+ 180.0+ 180.0+
#> [331]   1.0  180.0+ 180.0+ 171.0  180.0+ 174.0+ 180.0+ 180.0+  14.0+   7.0+
#> [341]   2.0  180.0+  36.0  180.0+   7.0    8.0+ 180.0+   2.0+ 180.0+  18.0 
#> [351] 180.0+ 180.0+ 180.0+   7.0    8.0+ 123.0  180.0+  51.0   19.0    1.0 
#> [361]  76.0  180.0+ 180.0+ 162.0    7.0+ 124.0    9.0  180.0+ 180.0+  12.0 
#> [371] 180.0+ 180.0+   2.0   76.0  180.0+ 173.0+ 180.0+ 180.0+ 180.0+  28.0 
#> [381] 180.0+ 180.0+  16.0+ 180.0+ 180.0+  16.0+ 180.0+   6.0  180.0+ 180.0+
#> [391]   7.0+  15.0    3.0+  13.0+ 180.0+   2.0    3.0+ 180.0+  20.0  180.0+
#> [401] 180.0+ 180.0+ 180.0+   3.0   87.0  180.0+   4.0+ 180.0+ 180.0+ 180.0+
#> [411] 180.0+ 180.0+   3.0  180.0+ 175.0    2.0   14.0+ 180.0+ 180.0+ 180.0+
#> [421] 179.0+ 180.0+ 180.0+ 159.0  180.0+  10.0    1.0   10.0   57.0  180.0+
#> [431]  11.0    3.0+   5.0  180.0+  12.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [441]  34.0  177.0+ 180.0+ 180.0+  11.0+   4.0+   7.0   15.0+ 180.0+  10.0 
#> [451] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   6.0  180.0+ 180.0+  85.0  180.0+
#> [461]  17.0+ 180.0+ 174.0+   7.0    0.5   12.0  180.0+  46.0  180.0+   4.0 
#> [471]   1.0  180.0+  52.0  180.0+ 180.0+ 180.0+ 180.0+   5.0  180.0+   1.0 
#> [481]  12.0  180.0+  79.0    3.0  168.0+ 176.0+ 180.0+  47.0   11.0    7.0 
#> [491] 180.0+  32.0  180.0+ 172.0  119.0   12.0  180.0+ 180.0+   1.0   80.0 
#> [501] 180.0+   2.0  180.0+  11.0    3.0   24.0   32.0   23.0    6.0  180.0+
#> [511]   3.0+ 180.0+   1.0   11.0   10.0+   3.0+   2.0+ 180.0+ 138.0  180.0+
#> [521]  71.0    8.0   10.0+ 180.0+ 164.0  118.0  173.0  180.0+  37.0  180.0+
#> [531]   7.0+  22.0    5.0+ 180.0+   3.0  166.0+  71.0   20.0+   1.0  180.0+
#> [541]  85.0    6.0+   6.0  180.0+   5.0    1.0  180.0+ 180.0+ 108.0  180.0+
#> [551] 180.0+   4.0    6.0  180.0+ 180.0+ 103.0  177.0+ 169.0   70.0    4.0 
#> [561] 180.0+ 180.0+   7.0+ 180.0+  20.0    8.0+  16.0  180.0+ 180.0+ 180.0+
#> [571] 177.0+   0.5    2.0   62.0  180.0+   4.0  180.0+  90.0  180.0+ 180.0+
#> [581] 180.0+ 180.0+   4.0    1.0   30.0  180.0+ 114.0  180.0+ 154.0    2.0 
#> [591] 180.0+   4.0+ 180.0+   1.0   12.0   16.0+ 180.0+ 180.0+  88.0  126.0 
#> [601] 180.0+ 165.0  180.0+ 180.0+   3.0+ 180.0+ 180.0+   4.0+   5.0  180.0+
#> [611] 180.0+ 180.0+  14.0    3.0  180.0+ 180.0+ 180.0+  38.0    4.0  177.0 
#> [621]   4.0+  20.0    4.0   22.0   65.0  180.0+ 115.0    3.0+   8.0+ 180.0+
#> [631]   4.0  119.0  110.0  180.0+ 180.0+  46.0   14.0  180.0+   8.0   25.0 
#> [641]  24.0   11.0  126.0  180.0+   4.0    1.0  178.0+  89.0   75.0    3.0+
#> [651]  33.0  158.0   74.0  180.0+ 168.0   52.0    4.0    4.0   50.0    1.0+
#> [661]   2.0   76.0  180.0+   8.0   26.0   53.0    7.0+   0.5   69.0  180.0+
#> 
#> $weights
#> NULL
#> 
#> $offset
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8326601