Cross-Validated GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::cv.glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.cv.glmnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.cv.glmnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.cv_glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

foldid	untyped	NULL		-

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nfolds	integer	10		\([3, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	TRUE	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	lambda.1se		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

standardize.response	logical	FALSE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.gaussian	character	-	covariance, naive	-

type.logistic	character	Newton	Newton, modified.Newton	-

type.measure	character	deviance	deviance, C	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.coxnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvCVGlmnet

Methods

Public methods

LearnerSurvCVGlmnet$new()

LearnerSurvCVGlmnet$selected_features()

LearnerSurvCVGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvCVGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvCVGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvCVGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.cv_glmnet")
print(learner)
#> 
#> ── <LearnerSurvCVGlmnet> (surv.cv_glmnet): Regularized Generalized Linear Model 
#> • Model: -
#> • Parameters: use_pred_offset=TRUE
#> • Packages: mlr3, mlr3proba, mlr3extralearners, and glmnet
#> • Predict Types: [crank], distr, and lp
#> • Feature Types: logical, integer, and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: offset, selected_features, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#> Measure: Partial Likelihood Deviance 
#> 
#>      Lambda Index Measure     SE Nonzero
#> min 0.00366    43   2.945 0.1622       6
#> 1se 0.07886    10   3.093 0.1446       3
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  32   5      1          0        1   121
#>   [2,]  33   2      0          2        0   150
#>   [3,]  35   5      1          2        0   172
#>   [4,]  35  10      1          9        0   106
#>   [5,]  34   5      0          5        0   120
#>   [6,]  36   1      0        180        1   155
#>   [7,]  38  12      1          8        1   120
#>   [8,]  36   5      1          0        1   115
#>   [9,]  38  16      1         10        0   160
#>  [10,]  38  12      1         11        1    92
#>  [11,]  40  12      1          9        0   153
#>  [12,]  42   3      1          1        1   130
#>  [13,]  42   2      0        180        1   100
#>  [14,]  40   6      0        180        1   138
#>  [15,]  40  11      1         10        1   120
#>  [16,]  42   2      0        180        0   100
#>  [17,]  43   3      1          0        1   100
#>  [18,]  43   4      1          0        1   130
#>  [19,]  41  13      1          1        0   140
#>  [20,]  45   9      1          7        0   110
#>  [21,]  41   5      1          4        1   141
#>  [22,]  43   2      0        180        1   140
#>  [23,]  45   2      0        180        1   140
#>  [24,]  46   2      1          1        0   126
#>  [25,]  44   3      1          0        1   180
#>  [26,]  46   7      1          2        0   166
#>  [27,]  43  29      0        180        1   180
#>  [28,]  45   4      1          0        0   124
#>  [29,]  47   6      1          0        1   116
#>  [30,]  46  13      1         10        0   100
#>  [31,]  47   4      1          3        1   160
#>  [32,]  43   3      1          0        1   124
#>  [33,]  45   8      1          0        1   117
#>  [34,]  45   5      0          5        0   141
#>  [35,]  46   2      1          1        1   122
#>  [36,]  46   6      1          0        1   100
#>  [37,]  44   4      1          0        1   114
#>  [38,]  44   9      1          8        1   135
#>  [39,]  45   5      0        180        1   190
#>  [40,]  46   5      1          3        0   130
#>  [41,]  46   4      0        180        1   121
#>  [42,]  44   2      0        180        0   142
#>  [43,]  46  15      0        180        1   120
#>  [44,]  48  12      1         11        0   200
#>  [45,]  47   5      1          3        1   130
#>  [46,]  47  10      0         10        1   140
#>  [47,]  50   1      1          0        1   129
#>  [48,]  48   2      1          0        0   184
#>  [49,]  50   4      1          1        0   125
#>  [50,]  50   6      1          2        1   140
#>  [51,]  49   7      1          7        1   110
#>  [52,]  46   9      1          9        1   122
#>  [53,]  50   7      0        180        1   110
#>  [54,]  49   2      0          2        0   105
#>  [55,]  51   1      0          1        1   145
#>  [56,]  49  15      1         11        1   160
#>  [57,]  47   2      0        180        0   150
#>  [58,]  49  23      0        179        1   112
#>  [59,]  46   6      1          0        1   156
#>  [60,]  52   2      0        180        1   170
#>  [61,]  51   3      1          2        0   113
#>  [62,]  50   1      1          0        0   150
#>  [63,]  50   9      0        180        0   130
#>  [64,]  49   7      1          4        1    90
#>  [65,]  47   8      0        180        0   160
#>  [66,]  51   8      0        180        1   140
#>  [67,]  52   2      0        180        0   155
#>  [68,]  46   1      1          1        0   145
#>  [69,]  50   4      1          1        0   150
#>  [70,]  48   7      1          0        1   110
#>  [71,]  53   8      0         36        1   160
#>  [72,]  49   9      1          3        0   102
#>  [73,]  49  15      0        180        1   160
#>  [74,]  53   5      0        180        1   140
#>  [75,]  53   5      0         77        0   159
#>  [76,]  53   7      1          0        0   199
#>  [77,]  54   6      1          3        0   129
#>  [78,]  51   3      1          1        0   140
#>  [79,]  50   2      0          5        1   106
#>  [80,]  50  14      1         13        0   170
#>  [81,]  53   8      1          7        0   160
#>  [82,]  51  25      1          1        0   202
#>  [83,]  49   5      1          2        1   150
#>  [84,]  53   4      0          4        0   140
#>  [85,]  48  11      1         10        0   120
#>  [86,]  51  13      0         99        1   160
#>  [87,]  54   9      1          0        1   138
#>  [88,]  55   3      1          1        0   150
#>  [89,]  52   7      1          2        0   154
#>  [90,]  54   9      1          1        0   130
#>  [91,]  55   4      1          2        0   150
#>  [92,]  51  13      1         11        0   145
#>  [93,]  50   5      1          4        1   150
#>  [94,]  54   4      1          0        1   121
#>  [95,]  52   4      0        180        0   183
#>  [96,]  55  28      1         13        1   160
#>  [97,]  49   6      1          0        1   130
#>  [98,]  49   1      0          1        1   110
#>  [99,]  50   7      1          1        0   156
#> [100,]  53   9      0          9        1    95
#> [101,]  53   8      1          0        1   130
#> [102,]  52   5      0        175        1   117
#> [103,]  55   2      0          2        0   145
#> [104,]  54   1      0        180        0   162
#> [105,]  54   7      1          0        1   100
#> [106,]  55   5      1          4        1   120
#> [107,]  52   8      0        180        0   119
#> [108,]  53  18      1          9        1   150
#> [109,]  54   3      0        180        1   180
#> [110,]  55   6      0        180        0   170
#> [111,]  52  16      0         16        0   152
#> [112,]  52  16      1         14        0   170
#> [113,]  53   4      0        180        1   150
#> [114,]  55   6      0        180        1   100
#> [115,]  55   6      1          5        1   138
#> [116,]  54  12      1          0        1   190
#> [117,]  54   3      0        180        0   128
#> [118,]  56   3      0          8        1   139
#> [119,]  55   1      0          2        0   130
#> [120,]  57   3      0          3        0   120
#> [121,]  54   7      1          2        0   129
#> [122,]  54   2      1          1        0   135
#> [123,]  57   5      1          3        1   138
#> [124,]  56   4      1          0        1   140
#> [125,]  52   2      0        180        0   140
#> [126,]  55  11      1          7        0   104
#> [127,]  52  15      1         14        0   130
#> [128,]  56  14      1         11        0   130
#> [129,]  53   3      1          0        1   200
#> [130,]  57  10      0        180        1   170
#> [131,]  58   8      0          8        1   130
#> [132,]  57   0      0          0        1   150
#> [133,]  53  21      1         13        1   130
#> [134,]  59   3      1          1        0   172
#> [135,]  58   6      1          0        1    90
#> [136,]  55   9      1          2        1   147
#> [137,]  55  13      0        166        1   140
#> [138,]  54  23      1          8        0   120
#> [139,]  57   4      1          2        1   185
#> [140,]  53   4      0        147        1   145
#> [141,]  55   5      0          5        1   131
#> [142,]  54   7      1          0        1   141
#> [143,]  56   4      0          4        0   164
#> [144,]  59  15      1         10        0   140
#> [145,]  58   9      1          0        1   180
#> [146,]  58   1      1          1        1   200
#> [147,]  57   1      0        180        0   148
#> [148,]  59   3      0        180        0   120
#> [149,]  58   4      1          0        1   160
#> [150,]  57   2      0          2        1   120
#> [151,]  57   5      0        180        1   130
#> [152,]  55   5      1          0        1   160
#> [153,]  57  10      1          9        0   103
#> [154,]  59   6      1          0        1   140
#> [155,]  59   5      0        180        1   155
#> [156,]  59   4      1          0        1   152
#> [157,]  61   9      0          9        1   160
#> [158,]  58   4      1          3        0   120
#> [159,]  60   0      1          0        1    80
#> [160,]  59   2      1          1        0   140
#> [161,]  58   8      0        161        1   140
#> [162,]  58  14      1          6        0   190
#> [163,]  61   4      1          3        0   151
#> [164,]  61   9      1          8        0   150
#> [165,]  61   3      1          2        1   102
#> [166,]  58   1      0          1        1   100
#> [167,]  61  20      1         13        0   130
#> [168,]  57  13      1         10        0   110
#> [169,]  57   2      1          0        1   116
#> [170,]  58  10      0         10        1   150
#> [171,]  57   4      1          3        0   138
#> [172,]  57  11      0        180        1   150
#> [173,]  61   3      0         17        0   143
#> [174,]  56  14      0         45        0   130
#> [175,]  56  13      1          6        1   158
#> [176,]  56  18      1         11        1   165
#> [177,]  59   9      1          0        1    80
#> [178,]  58  11      0        172        1   135
#> [179,]  60  12      1          0        1   114
#> [180,]  55   9      1          7        1   135
#> [181,]  61   4      1          0        1   115
#> [182,]  56   8      1          8        0   120
#> [183,]  61  13      1         12        1   130
#> [184,]  59  11      1          8        1   190
#> [185,]  58   5      1          1        1   135
#> [186,]  61   8      0         77        0   120
#> [187,]  58   8      1          5        0   152
#> [188,]  62  10      1          0        1   153
#> [189,]  62   7      1          2        1   180
#> [190,]  57   3      1          0        0   100
#> [191,]  61  18      0        170        0   140
#> [192,]  61  28      1          7        0   133
#> [193,]  57   7      0        169        0   180
#> [194,]  61   7      0          7        1   150
#> [195,]  60   7      0          7        0   147
#> [196,]  61   6      0          6        0   134
#> [197,]  62   4      1          0        0   160
#> [198,]  60  17      1          8        1   140
#> [199,]  58   3      1          0        1   146
#> [200,]  62   4      1          3        0   173
#> [201,]  58   2      0         30        0   202
#> [202,]  59   1      0        180        0   155
#> [203,]  59  16      1          9        1   133
#> [204,]  63   6      0         28        1   120
#> [205,]  61   5      0          5        1   110
#> [206,]  57   2      1          1        0   159
#> [207,]  62  17      1         10        1   180
#> [208,]  62   1      1          0        1   172
#> [209,]  58   7      0        180        1   150
#> [210,]  63   1      0        180        1   130
#> [211,]  63   4      1          3        0   222
#> [212,]  62   3      0        180        1   105
#> [213,]  63   4      0        180        1   190
#> [214,]  63  15      1         10        1   126
#> [215,]  64   4      0        180        0   130
#> [216,]  63   4      1          1        0   155
#> [217,]  60  18      1         13        0   132
#> [218,]  59   8      0        180        1   140
#> [219,]  61   9      1          9        1   150
#> [220,]  58   9      1          9        0   110
#> [221,]  62   7      0          7        0   150
#> [222,]  59   1      0         22        1   162
#> [223,]  58   2      0        180        0   127
#> [224,]  59   4      0        180        0   196
#> [225,]  59   5      1          1        0   148
#> [226,]  60   7      1          1        1    90
#> [227,]  59   4      0          4        0   149
#> [228,]  60   3      0          3        0   168
#> [229,]  64  10      1          9        0   160
#> [230,]  60   8      0         17        1   130
#> [231,]  61   6      1          1        1   117
#> [232,]  63  10      1          0        1   148
#> [233,]  63  14      1          9        0   123
#> [234,]  65  36      1         11        0   140
#> [235,]  63   4      1          3        0   162
#> [236,]  66   3      1          1        0   127
#> [237,]  63  12      1          9        0   114
#> [238,]  66   5      1          0        1   110
#> [239,]  65   8      1          0        0   168
#> [240,]  64   0      0          0        1    90
#> [241,]  60   6      0        180        0   130
#> [242,]  64  21      1         10        0   190
#> [243,]  64   9      0        180        0   150
#> [244,]  61   4      0        180        1   113
#> [245,]  65   3      0        180        1   190
#> [246,]  63  16      1          7        1   110
#> [247,]  64   7      0        180        1   120
#> [248,]  66   6      1          1        1   130
#> [249,]  63  12      0         12        1   150
#> [250,]  62   3      1          1        1   199
#> [251,]  65   6      0          9        0   112
#> [252,]  65   3      1          0        1    80
#> [253,]  63   5      1          4        0   170
#> [254,]  63   2      1          1        0   180
#> [255,]  62  13      1         11        0   180
#> [256,]  66  16      1         11        1   169
#> [257,]  61  14      1          5        0   140
#> [258,]  61  15      1         10        0   130
#> [259,]  63   9      1          8        1   160
#> [260,]  63   3      1          2        0   120
#> [261,]  64  19      1          8        1   160
#> [262,]  65   8      1          0        1   140
#> [263,]  67   6      0        180        1   170
#> [264,]  65  15      1         11        1   160
#> [265,]  64  13      1         12        1   150
#> [266,]  64   6      1          0        1   125
#> [267,]  66   7      1          0        1   115
#> [268,]  66  13      1          0        0   118
#> [269,]  65   3      0          3        0   105
#> [270,]  66   3      1          0        1   135
#> [271,]  66   6      1          0        1   140
#> [272,]  65   2      1          1        1   170
#> [273,]  68   1      0        180        1   166
#> [274,]  64  10      1          9        1   110
#> [275,]  63   7      1          0        0   162
#> [276,]  67   8      1          1        1   130
#> [277,]  63  10      0         16        1   160
#> [278,]  66  14      0        180        0   130
#> [279,]  63   8      1          1        1   162
#> [280,]  63   1      1          0        1   155
#> [281,]  63  10      0         18        1   130
#> [282,]  67  11      0         11        0   150
#> [283,]  68  11      0        180        0   160
#> [284,]  69  12      0         15        1   140
#> [285,]  66  15      1         13        1   160
#> [286,]  63   2      0        180        0   150
#> [287,]  65  11      1          6        0   130
#> [288,]  69  21      1         10        0   180
#> [289,]  66   9      1          8        0   130
#> [290,]  68  14      1         13        1   140
#> [291,]  65   8      1          0        1    90
#> [292,]  69   1      1          0        0   170
#> [293,]  67   1      0        180        1   160
#> [294,]  65   1      1          0        0   133
#> [295,]  67   7      1          4        1   130
#> [296,]  65   6      0          6        0    80
#> [297,]  65  10      1          1        1   148
#> [298,]  66  19      1         12        1   150
#> [299,]  69   6      0         99        1   140
#> [300,]  64   4      0        179        0   160
#> [301,]  70  15      1         12        1   132
#> [302,]  64  11      0         11        0   125
#> [303,]  64   4      0        180        1   140
#> [304,]  64   0      1          0        1   118
#> [305,]  67   2      0         18        0   131
#> [306,]  68   4      1          0        1   160
#> [307,]  69   4      1          3        1   150
#> [308,]  65  13      1         12        1   130
#> [309,]  64  21      0         21        1   155
#> [310,]  65   1      0          1        1   120
#> [311,]  68  18      1          0        1   160
#> [312,]  65   6      0        101        1   115
#> [313,]  68   4      0          4        1   190
#> [314,]  70   7      1          0        1   190
#> [315,]  68   7      0        150        0   210
#> [316,]  71  20      1          0        1   160
#> [317,]  66   9      1          3        1   151
#> [318,]  66   1      1          1        1   165
#> [319,]  70   4      1          0        1   180
#> [320,]  69   8      0        180        1   153
#> [321,]  70  14      0        171        0   166
#> [322,]  67  10      1          9        0   200
#> [323,]  67   6      1          4        0   130
#> [324,]  69   0      0          0        1   148
#> [325,]  68   7      1          0        1   150
#> [326,]  69   3      1          2        0   151
#> [327,]  71   7      0          7        0   230
#> [328,]  71   6      0         45        1   158
#> [329,]  69   5      0          5        1   142
#> [330,]  71   3      0        103        0   133
#> [331,]  69   3      0          3        1   130
#> [332,]  70  22      1         13        0   103
#> [333,]  67   1      0         36        1   104
#> [334,]  67   5      0          5        0   130
#> [335,]  68   6      0        180        0   145
#> [336,]  69   6      1          4        1   174
#> [337,]  72   3      1          0        1   132
#> [338,]  72   7      0          7        1   110
#> [339,]  69   8      1          7        1   108
#> [340,]  67   3      0        180        0   110
#> [341,]  68  18      0         18        1   100
#> [342,]  67  14      0        172        1   140
#> [343,]  69  11      1          0        1   120
#> [344,]  66   2      0        180        0   130
#> [345,]  67   7      1          4        0   122
#> [346,]  69   4      1          3        0   132
#> [347,]  67  13      1          9        0   130
#> [348,]  70   3      0        123        0   130
#> [349,]  69   1      0          1        1   110
#> [350,]  67   1      0          1        1    60
#> [351,]  67   4      0         60        1   136
#> [352,]  69   5      0         76        0   120
#> [353,]  72  13      1         11        1   195
#> [354,]  68  10      1          8        1   160
#> [355,]  66  24      1         13        0   130
#> [356,]  70  35      1          0        1   105
#> [357,]  70   7      0          7        0   102
#> [358,]  68   7      1          2        0   135
#> [359,]  69  10      1          6        1   120
#> [360,]  70  11      0        180        1   210
#> [361,]  72  19      1          8        0   120
#> [362,]  72  12      1         10        0   170
#> [363,]  67   8      0        180        1   170
#> [364,]  67   9      0        180        0   158
#> [365,]  72   2      0          2        1   100
#> [366,]  67   4      1          1        0   134
#> [367,]  72   6      1          5        0   115
#> [368,]  69   3      0        180        0   220
#> [369,]  72   5      0         28        0   120
#> [370,]  71   5      0        180        0   191
#> [371,]  73   6      0        180        1   117
#> [372,]  69   8      1          1        0   164
#> [373,]  68   7      0        180        1   130
#> [374,]  70   4      0        180        0   180
#> [375,]  73   6      1          0        1   270
#> [376,]  72   8      1          1        1   150
#> [377,]  71   2      1          0        1   180
#> [378,]  73   7      0          7        1   140
#> [379,]  68  15      1         13        1   130
#> [380,]  72   6      0        180        1   130
#> [381,]  73   0      0        180        1   161
#> [382,]  74   8      1          0        1    85
#> [383,]  69   2      1          0        1   110
#> [384,]  71   3      1          1        0   150
#> [385,]  71  15      1         11        0   165
#> [386,]  74  20      0         20        1   180
#> [387,]  68   9      0        180        1   120
#> [388,]  74   0      1          0        1    90
#> [389,]  73   3      1          0        1   136
#> [390,]  71  17      1         11        0   160
#> [391,]  71   3      1          2        1   190
#> [392,]  73  10      1          8        0   106
#> [393,]  70  26      1         11        1   120
#> [394,]  73   4      0         58        1   160
#> [395,]  72   5      1          3        1   160
#> [396,]  70   3      0        180        1   154
#> [397,]  73   6      0        180        0   110
#> [398,]  71   7      1          2        0   143
#> [399,]  72   8      1          0        1   140
#> [400,]  74   3      0          3        1   150
#> [401,]  73  17      1         11        0   140
#> [402,]  71  13      1          8        0   121
#> [403,]  69   2      1          1        1    80
#> [404,]  71  14      1         13        1   170
#> [405,]  74   7      1          0        1   117
#> [406,]  72  10      1          8        1   153
#> [407,]  72  15      1         13        0   156
#> [408,]  71  10      1          9        1   120
#> [409,]  73  10      1          9        1   146
#> [410,]  72  10      1          9        1   160
#> [411,]  73   1      0          1        1    80
#> [412,]  75   9      1          7        0   140
#> [413,]  75  13      1          1        1   130
#> [414,]  71  11      1          8        0   110
#> [415,]  71   4      0          4        0   134
#> [416,]  72  15      1         12        1   120
#> [417,]  70   7      1          4        0   184
#> [418,]  72   1      1          1        0   168
#> [419,]  73  10      0        180        0   162
#> [420,]  72  11      0         11        1   140
#> [421,]  73  12      1         12        1   140
#> [422,]  72   4      1          0        1   197
#> [423,]  71   3      1          0        0   144
#> [424,]  73   5      0        180        0   126
#> [425,]  74  34      1          8        1   233
#> [426,]  71  32      1         12        1   107
#> [427,]  72   5      0        180        0   154
#> [428,]  76   5      0          5        1   130
#> [429,]  77  11      0         11        1   150
#> [430,]  77   4      0          4        0   185
#> [431,]  72   7      1          2        0   142
#> [432,]  74   3      0          3        1   128
#> [433,]  74   2      1          1        0   140
#> [434,]  76   8      1          0        1   141
#> [435,]  75  23      1         14        1   110
#> [436,]  74   2      0        180        0   190
#> [437,]  72   4      0         85        1   120
#> [438,]  72   4      1          3        0   160
#> [439,]  76  17      1          0        1   200
#> [440,]  76  13      1         10        0   110
#> [441,]  75   4      1          0        1   122
#> [442,]  75   7      0          7        0   190
#> [443,]  75   0      0          0        1   130
#> [444,]  73  13      1         11        0   195
#> [445,]  75  12      0         12        1   160
#> [446,]  74   8      1          0        1   105
#> [447,]  76  13      1          8        1   148
#> [448,]  75   4      1          2        1   188
#> [449,]  74   6      0        180        0   160
#> [450,]  76   4      0          4        1   155
#> [451,]  73   1      0         52        1   105
#> [452,]  73   0      0        180        0   156
#> [453,]  78  12      1         11        1   160
#> [454,]  76   5      0        180        0   185
#> [455,]  74  10      1          0        1   135
#> [456,]  74   8      1          8        1   170
#> [457,]  75   9      0        180        1   140
#> [458,]  73  33      1         12        1   175
#> [459,]  77   5      1          0        0   123
#> [460,]  77  12      1          9        1   100
#> [461,]  73  10      1          9        0   146
#> [462,]  77   1      1          0        1    90
#> [463,]  73   7      1          0        0   174
#> [464,]  75   3      1          1        1   171
#> [465,]  74   9      1          8        0   200
#> [466,]  79  10      1          8        0   190
#> [467,]  74   2      1          0        1   130
#> [468,]  78  18      0         18        1   144
#> [469,]  77   3      0        180        0   110
#> [470,]  76  29      0         47        0    90
#> [471,]  73   8      1          1        1   162
#> [472,]  73  11      1          2        1   110
#> [473,]  74  15      0        180        1   172
#> [474,]  74   7      0          7        0   161
#> [475,]  78  32      1          9        1   198
#> [476,]  80  10      1          6        1   147
#> [477,]  78   0      0        180        1   212
#> [478,]  78  13      1          5        0   130
#> [479,]  75   5      0        119        1   150
#> [480,]  75  12      1          1        1   120
#> [481,]  80   8      0          8        1   120
#> [482,]  75  13      1          6        0   150
#> [483,]  76   1      0          1        1    83
#> [484,]  79   4      0         80        0   145
#> [485,]  78  12      1          9        0   150
#> [486,]  78   2      1          1        0   130
#> [487,]  77   2      1          0        1   143
#> [488,]  78  10      0        180        1   130
#> [489,]  76  11      1          0        0   120
#> [490,]  75  11      1          4        0   162
#> [491,]  76   7      0         29        1   150
#> [492,]  77  24      0         24        1   160
#> [493,]  79   8      0         32        1   120
#> [494,]  80   6      0          6        1   150
#> [495,]  78   6      1          0        1   240
#> [496,]  76   3      1          0        1   140
#> [497,]  78  11      1          1        1   140
#> [498,]  79  11      0        180        0   160
#> [499,]  79   2      1          0        1   121
#> [500,]  78  14      1          0        1   140
#> [501,]  81   1      0          1        0   130
#> [502,]  76   4      0          4        1   160
#> [503,]  79   4      0          4        1   125
#> [504,]  76  10      1          8        0   180
#> [505,]  76  12      1         10        1   127
#> [506,]  77   6      0          6        1   107
#> [507,]  80   3      1          0        1   120
#> [508,]  75   2      1          1        1   204
#> [509,]  76   1      0          1        1   140
#> [510,]  77  31      1          3        1   161
#> [511,]  76   1      0          1        1    90
#> [512,]  78   7      1          0        1   110
#> [513,]  77   7      0        180        1   170
#> [514,]  79   4      1          0        1   120
#> [515,]  77   9      1          4        0   141
#> [516,]  82   5      0          8        1   120
#> [517,]  78   4      0         59        1   112
#> [518,]  76   7      0        161        0   151
#> [519,]  79  10      0         10        1   120
#> [520,]  80  15      1          0        1    90
#> [521,]  78  32      0        180        1   130
#> [522,]  79   1      0         37        1   140
#> [523,]  81   3      0        180        0   184
#> [524,]  78   7      0          7        1   147
#> [525,]  77  13      1          0        1   190
#> [526,]  78  15      0         15        0   165
#> [527,]  80   5      1          1        1   108
#> [528,]  78  26      1          5        0   194
#> [529,]  76   1      0        166        0   131
#> [530,]  78  20      1          0        1   109
#> [531,]  80   1      0          1        0   100
#> [532,]  77  10      1          8        1   130
#> [533,]  82   3      1          1        1   144
#> [534,]  77   5      0         85        0   188
#> [535,]  80   2      1          1        0   168
#> [536,]  80   6      1          0        1   119
#> [537,]  78   2      0        180        0   148
#> [538,]  80   5      0          5        1   130
#> [539,]  82   1      1          0        1    82
#> [540,]  79  10      0        180        1   150
#> [541,]  77   4      0        180        1    98
#> [542,]  81   1      0        108        0   129
#> [543,]  78  12      0        180        0   134
#> [544,]  79   1      0        125        0   193
#> [545,]  82  21      1          2        0   155
#> [546,]  84  22      1         10        0   180
#> [547,]  79   4      0          4        1   121
#> [548,]  82   5      0        180        0   110
#> [549,]  83   5      0        180        0   148
#> [550,]  79   7      1          6        0   130
#> [551,]  83   4      0        103        0    97
#> [552,]  81  11      1          8        0   160
#> [553,]  80  11      1          8        0   170
#> [554,]  79   4      0          4        1   183
#> [555,]  79   1      0        180        1   170
#> [556,]  81  15      0        180        1   140
#> [557,]  80   7      1          0        1   146
#> [558,]  84   5      1          1        1    85
#> [559,]  81  20      1          9        0   170
#> [560,]  83   8      0          8        0   115
#> [561,]  80  11      1          8        0   110
#> [562,]  85   4      0        180        0    90
#> [563,]  81   2      1          1        0   198
#> [564,]  83   2      0          2        1   155
#> [565,]  82   6      0        128        1   100
#> [566,]  84   4      0        167        0   198
#> [567,]  82  23      1          0        0   110
#> [568,]  84   5      0        180        1   203
#> [569,]  84   1      0         38        1   205
#> [570,]  81   4      0         90        1   138
#> [571,]  80  13      1          8        1   140
#> [572,]  84   4      0         89        1   129
#> [573,]  80   2      1          0        1   130
#> [574,]  79   4      0          4        1    60
#> [575,]  80   6      0         71        1   189
#> [576,]  82  19      0         19        0   120
#> [577,]  80  30      1         13        0   220
#> [578,]  79  14      1          0        0   110
#> [579,]  83   3      0        114        0    98
#> [580,]  83   2      0        154        0   130
#> [581,]  82   0      0          2        1   100
#> [582,]  85   9      1          6        1   160
#> [583,]  83   1      0        180        0   160
#> [584,]  81   4      0          4        0   175
#> [585,]  84  15      1         13        1   110
#> [586,]  81   1      0          1        1   145
#> [587,]  81  12      0         12        1   163
#> [588,]  82   5      1          0        1   146
#> [589,]  86  12      0        180        1   120
#> [590,]  83  12      1          2        1   170
#> [591,]  81  19      1         14        0   120
#> [592,]  82   3      1          2        0   130
#> [593,]  86   8      0          8        1   132
#> [594,]  81  16      1          9        0   180
#> [595,]  86   3      0          3        1   140
#> [596,]  82   9      0        180        1   134
#> [597,]  84   3      0        180        1   120
#> [598,]  81  13      0        180        0   152
#> [599,]  85   3      0          3        1   118
#> [600,]  81   2      1          0        1   118
#> [601,]  83   9      0        180        1   149
#> [602,]  82   1      0        180        1   193
#> [603,]  83   4      0          4        0   130
#> [604,]  87   2      0          5        1   137
#> [605,]  82  14      1         11        1   103
#> [606,]  86   6      1          0        1   140
#> [607,]  84  16      0         70        1   150
#> [608,]  83  19      0         43        0   150
#> [609,]  84   3      1          2        0   125
#> [610,]  88  14      1          3        1   130
#> [611,]  84   3      0          3        1   121
#> [612,]  87   2      0        180        0   113
#> [613,]  84   9      0         92        1   110
#> [614,]  84   3      0        180        1   170
#> [615,]  86   4      0         38        1   122
#> [616,]  86   6      0          6        1   117
#> [617,]  84  13      0         62        1   100
#> [618,]  86   6      1          1        0   112
#> [619,]  83  20      1          3        1   150
#> [620,]  88   4      0          4        1   115
#> [621,]  86   6      0         46        0   173
#> [622,]  88   3      0        115        0   110
#> [623,]  83   3      0          3        1   130
#> [624,]  87   8      0          8        1   157
#> [625,]  86  15      1          8        1   109
#> [626,]  89   5      0        119        1   140
#> [627,]  87   1      0          1        0   170
#> [628,]  84   8      0        180        1   119
#> [629,]  85   8      0          8        1   136
#> [630,]  87  29      0         29        1    97
#> [631,]  87  15      1          9        1   138
#> [632,]  89  10      0         46        1   170
#> [633,]  88   1      0          1        0   135
#> [634,]  86   4      0        180        1   145
#> [635,]  91   8      0          8        0   100
#> [636,]  87   6      1          0        0   125
#> [637,]  90  11      1         10        1   186
#> [638,]  86   9      1          7        0   130
#> [639,]  90   4      1          0        0   121
#> [640,]  91   1      0          1        1    74
#> [641,]  88   3      1          2        0   159
#> [642,]  89   3      1          1        1   160
#> [643,]  92   1      0          1        1   167
#> [644,]  91   3      0         33        1   137
#> [645,]  88   5      0        158        0   100
#> [646,]  89  12      1          0        1   130
#> [647,]  89   2      0        168        0   118
#> [648,]  92   7      0          7        1   110
#> [649,]  91   0      0          0        0     0
#> [650,]  89  14      0        180        1    84
#> [651,]  90  18      0        180        0   188
#> [652,]  90  19      1         11        1   129
#> [653,]  94   6      0         50        0    78
#> [654,]  91   2      0          2        1   116
#> [655,]  93   8      0        179        1   110
#> [656,]  94   8      0          8        1   142
#> [657,]  92   4      0         76        1   149
#> [658,]  91   1      0        180        0   158
#> [659,]  90  16      0         16        1   106
#> [660,]  90   3      0         67        0   162
#> [661,]  96   3      0         12        1    97
#> [662,]  94   3      0         26        1   144
#> [663,]  91  12      0         53        1   212
#> [664,]  91   7      0          7        0   135
#> [665,]  93   0      1          0        1   122
#> [666,]  92   5      0         69        0   139
#> [667,]  92   2      0          2        0   112
#> [668,]  93   4      0        180        1   135
#> [669,]  96   3      1          0        1   104
#> [670,]  96  15      1          0        1   140
#> 
#> $y
#>   [1]   5.0+   2.0+   5.0+ 180.0+   5.0+ 180.0+  12.0    5.0+ 180.0+ 180.0+
#>  [11] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   3.0  180.0+ 180.0+ 180.0+
#>  [21]   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   6.0+ 180.0+
#>  [31] 180.0+ 180.0+ 180.0+   5.0+ 161.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0+
#>  [41] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  10.0+ 172.0+ 180.0+ 180.0+ 180.0+
#>  [51]   7.0  180.0+ 180.0+   2.0    1.0  179.0+ 180.0+ 179.0+ 180.0+ 180.0+
#>  [61] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   7.0 
#>  [71]  36.0  180.0+ 180.0+ 180.0+  77.0  180.0+ 180.0+ 180.0+   5.0  180.0+
#>  [81] 180.0+ 180.0+ 180.0+   4.0+ 180.0+  99.0  180.0+ 180.0+   7.0+ 180.0+
#>  [91] 180.0+  13.0+ 171.0+ 180.0+ 180.0+  28.0    6.0+   1.0  180.0+   9.0+
#> [101] 180.0+ 175.0+   2.0  180.0+   7.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [111]  16.0+  16.0  180.0+ 180.0+ 180.0+  12.0+ 180.0+   8.0    2.0    3.0+
#> [121] 180.0+ 180.0+ 140.0  165.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [131]   8.0+   0.5  180.0+ 180.0+ 180.0+  15.0  166.0+ 180.0+   4.0+ 147.0+
#> [141]   5.0+ 180.0+   4.0+ 180.0+   9.0+   1.0  180.0+ 180.0+ 180.0+   2.0 
#> [151] 180.0+ 180.0+ 180.0+  64.0  180.0+ 180.0+   9.0+ 180.0+   0.5  180.0+
#> [161] 161.0+ 171.0+ 180.0+ 180.0+   3.0    1.0  180.0+ 180.0+ 180.0+  10.0+
#> [171] 180.0+ 180.0+  17.0   45.0  180.0+ 180.0+   9.0+ 172.0+ 172.0+  24.0 
#> [181] 180.0+   8.0  180.0+ 180.0+ 180.0+  77.0    8.0+ 180.0+ 180.0+ 180.0+
#> [191] 170.0   94.0  169.0    7.0    7.0+   6.0  180.0+ 180.0+   3.0+ 180.0+
#> [201]  30.0  180.0+ 180.0+  28.0    5.0  180.0+ 180.0+   1.0  180.0+ 180.0+
#> [211] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+ 180.0+ 180.0+   9.0 
#> [221]   7.0+  22.0  180.0+ 180.0+ 180.0+ 180.0+   4.0+   3.0+ 167.0   17.0 
#> [231] 180.0+ 180.0+  14.0+  36.0  180.0+   3.0+  12.0  180.0+ 180.0+   0.5 
#> [241] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  12.0  180.0+
#> [251]   9.0    3.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   3.0+
#> [261] 103.0   15.0  180.0+ 180.0+  13.0  180.0+ 179.0+ 166.0+   3.0    3.0+
#> [271] 180.0+ 175.0+ 180.0+ 180.0+   7.0+   8.0   16.0  180.0+ 180.0+   1.0+
#> [281]  18.0   11.0+ 180.0+  15.0  180.0+ 180.0+ 180.0+ 174.0+ 180.0+ 180.0+
#> [291]   8.0+ 175.0  180.0+ 180.0+ 180.0+   6.0  180.0+  19.0+  99.0  179.0+
#> [301] 180.0+  11.0+ 180.0+   0.5   18.0  180.0+ 152.0+ 180.0+  21.0+   1.0 
#> [311]  18.0+ 101.0    4.0    7.0+ 150.0  180.0+ 180.0+   1.0  180.0+ 180.0+
#> [321] 171.0  174.0+   6.0    0.5  180.0+ 180.0+   7.0+  45.0    5.0+ 103.0 
#> [331]   3.0+ 180.0+  36.0    5.0+ 180.0+  97.0  180.0+   7.0    8.0+ 180.0+
#> [341]  18.0  172.0+ 180.0+ 180.0+   7.0  180.0+  13.0+ 123.0    1.0    1.0 
#> [351]  60.0   76.0  132.0   10.0+ 180.0+ 180.0+   7.0+   7.0+ 180.0+ 180.0+
#> [361] 180.0+  12.0  180.0+ 180.0+   2.0   76.0  180.0+ 180.0+  28.0  180.0+
#> [371] 180.0+ 180.0+ 180.0+ 180.0+   6.0  180.0+ 180.0+   7.0+  15.0  180.0+
#> [381] 180.0+ 180.0+   2.0    3.0+ 180.0+  20.0  180.0+   0.5  180.0+ 180.0+
#> [391]   3.0   87.0  180.0+  58.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+   3.0 
#> [401] 180.0+ 175.0    2.0   14.0+ 180.0+  10.0+ 180.0+ 179.0+ 180.0+ 159.0 
#> [411]   1.0  180.0+  13.0  180.0+   4.0+ 180.0+ 104.0+   1.0  180.0+  11.0 
#> [421]  12.0  180.0+ 180.0+ 180.0+  34.0  177.0+ 180.0+   5.0   11.0+   4.0+
#> [431]   7.0    3.0  180.0+ 180.0+ 180.0+ 180.0+  85.0  180.0+  17.0+ 174.0+
#> [441]   4.0    7.0    0.5  180.0+  12.0  180.0+ 180.0+  46.0  180.0+   4.0 
#> [451]  52.0  180.0+  12.0  180.0+ 180.0+   8.0  180.0+  33.0    5.0  180.0+
#> [461] 180.0+   1.0    7.0+   3.0  168.0+ 180.0+ 176.0+  18.0  180.0+  47.0 
#> [471] 180.0+  11.0  180.0+   7.0   32.0   10.0  180.0+ 172.0  119.0   12.0 
#> [481]   8.0  180.0+   1.0   80.0  180.0+ 180.0+   2.0  180.0+  11.0  152.0+
#> [491]  29.0   24.0   32.0    6.0  180.0+   3.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [501]   1.0    4.0    4.0   10.0+ 180.0+   6.0    3.0+   2.0+   1.0  171.0 
#> [511]   1.0   43.0  180.0+ 138.0   71.0    8.0   59.0  161.0   10.0+ 180.0+
#> [521] 180.0+  37.0  180.0+   7.0+  22.0   15.0+   5.0+ 171.0+ 166.0+  20.0+
#> [531]   1.0   10.0  180.0+  85.0   10.0    6.0  180.0+   5.0    1.0  180.0+
#> [541] 180.0+ 108.0  180.0+ 125.0  180.0+ 180.0+   4.0  180.0+ 180.0+ 180.0+
#> [551] 103.0  180.0+ 169.0    4.0  180.0+ 180.0+   7.0+ 180.0+  20.0    8.0+
#> [561] 180.0+ 180.0+ 180.0+   2.0  128.0  167.0   62.0  180.0+  38.0   90.0 
#> [571] 180.0+  89.0  180.0+   4.0   71.0   19.0   30.0  180.0+ 114.0  154.0 
#> [581]   2.0  180.0+ 180.0+   4.0+ 180.0+   1.0   12.0    5.0+ 180.0+  77.0 
#> [591] 180.0+   3.0    8.0  180.0+   3.0  180.0+ 180.0+ 180.0+   3.0+ 180.0+
#> [601] 180.0+ 180.0+   4.0+   5.0  174.0    6.0   70.0   43.0  180.0+  14.0 
#> [611]   3.0  180.0+  92.0  180.0+  38.0    6.0+  62.0    6.0+  20.0    4.0 
#> [621]  46.0  115.0    3.0+   8.0+ 180.0+ 119.0    1.0+ 180.0+   8.0   29.0 
#> [631] 180.0+  46.0    1.0+ 180.0+   8.0   25.0   11.0  180.0+   4.0    1.0 
#> [641]  75.0    3.0+   1.0   33.0  158.0  180.0+ 168.0    7.0    0.5  180.0+
#> [651] 180.0+ 180.0+  50.0    2.0  179.0+   8.0+  76.0  180.0+  16.0   67.0 
#> [661]  12.0   26.0   53.0    7.0+   0.5   69.0    2.0  180.0+   3.0   15.0+
#> 
#> $weights
#> NULL
#> 
#> $offset
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8665788