Cross-Validated GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::cv.glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.cv.glmnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.cv.glmnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.cv_glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

foldid	untyped	NULL		-

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nfolds	integer	10		\([3, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	TRUE	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	lambda.1se		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

standardize.response	logical	FALSE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.gaussian	character	-	covariance, naive	-

type.logistic	character	Newton	Newton, modified.Newton	-

type.measure	character	deviance	deviance, C	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.coxnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvCVGlmnet

Methods

Public methods

LearnerSurvCVGlmnet$new()

LearnerSurvCVGlmnet$selected_features()

LearnerSurvCVGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvCVGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvCVGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvCVGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("surv.cv_glmnet")
print(learner)
#> 
#> ── <LearnerSurvCVGlmnet> (surv.cv_glmnet): Regularized Generalized Linear Model 
#> • Model: -
#> • Parameters: use_pred_offset=TRUE
#> • Packages: mlr3, mlr3proba, mlr3extralearners, and glmnet
#> • Predict Types: [crank], distr, and lp
#> • Feature Types: logical, integer, and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: offset, selected_features, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("grace")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#> Measure: Partial Likelihood Deviance 
#> 
#>      Lambda Index Measure     SE Nonzero
#> min 0.00395    43   2.794 0.1487       6
#> 1se 0.07057    12   2.923 0.1351       4
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  28   9      0        180        1   107
#>   [2,]  32   5      1          0        1   121
#>   [3,]  35   5      1          2        0   172
#>   [4,]  34   5      0          5        0   120
#>   [5,]  35   2      0        180        0   121
#>   [6,]  37   9      0        180        1   151
#>   [7,]  38  13      1          0        1   161
#>   [8,]  36   1      0        180        1   155
#>   [9,]  35   0      0        180        1   119
#>  [10,]  38  12      1          8        1   120
#>  [11,]  36   5      1          0        1   115
#>  [12,]  38  16      1         10        0   160
#>  [13,]  40  12      1          9        0   153
#>  [14,]  42   3      1          1        1   130
#>  [15,]  42   2      0          2        0   140
#>  [16,]  40   6      0        180        1   138
#>  [17,]  40  11      1         10        1   120
#>  [18,]  42   2      0        180        0   100
#>  [19,]  43   3      1          0        1   100
#>  [20,]  41   2      1          1        0   166
#>  [21,]  40   1      1          0        1   145
#>  [22,]  43   4      1          0        1   130
#>  [23,]  42   4      0        180        0   162
#>  [24,]  42  15      1         13        1   125
#>  [25,]  40   3      1          1        0   170
#>  [26,]  43   2      1          1        1   116
#>  [27,]  42   2      0        180        1   124
#>  [28,]  44   5      1          1        0   170
#>  [29,]  45   9      1          7        0   110
#>  [30,]  45   6      0        180        1   170
#>  [31,]  41   5      1          4        1   141
#>  [32,]  45   2      0        180        1   140
#>  [33,]  46  15      0        180        0   120
#>  [34,]  47   4      1          3        0   118
#>  [35,]  48  15      0        180        1   160
#>  [36,]  45   3      0        150        0   130
#>  [37,]  46   7      1          2        0   166
#>  [38,]  43  10      0        180        0   185
#>  [39,]  47   6      1          0        1   116
#>  [40,]  45   8      1          0        1   117
#>  [41,]  46   2      1          1        1   122
#>  [42,]  44   4      1          0        1   114
#>  [43,]  47   2      0        180        0   108
#>  [44,]  44   9      1          8        1   135
#>  [45,]  45   5      0        180        1   190
#>  [46,]  46   4      0        180        1   121
#>  [47,]  44   2      0        180        0   142
#>  [48,]  45   9      1          0        1   145
#>  [49,]  47   3      1          1        1   120
#>  [50,]  48   3      0        180        0   154
#>  [51,]  48  12      1         11        0   200
#>  [52,]  47   5      1          3        1   130
#>  [53,]  49   4      0        180        0   117
#>  [54,]  47  10      0         10        1   140
#>  [55,]  48   2      1          0        0   184
#>  [56,]  47   7      0        180        0   145
#>  [57,]  50   4      1          1        0   125
#>  [58,]  50   6      1          2        1   140
#>  [59,]  46   3      1          1        1   140
#>  [60,]  50   7      0        180        1   110
#>  [61,]  49   2      0          2        0   105
#>  [62,]  49  15      1         11        1   160
#>  [63,]  49  23      0        179        1   112
#>  [64,]  46   6      1          0        1   156
#>  [65,]  52   2      0        180        1   170
#>  [66,]  50   7      1          0        1    92
#>  [67,]  50   4      0          4        1   100
#>  [68,]  51   3      1          2        0   113
#>  [69,]  50   9      0        180        0   130
#>  [70,]  49   7      1          4        1    90
#>  [71,]  47   8      0        180        0   160
#>  [72,]  51   8      0        180        1   140
#>  [73,]  46   3      0        180        1   120
#>  [74,]  50   4      1          1        0   150
#>  [75,]  48  17      1         10        0   111
#>  [76,]  47   2      1          1        0   110
#>  [77,]  52   4      1          4        0   152
#>  [78,]  49   9      1          3        0   102
#>  [79,]  54  17      1         12        1   102
#>  [80,]  53   7      1          0        0   199
#>  [81,]  51   3      1          1        0   140
#>  [82,]  50   2      0          5        1   106
#>  [83,]  50  10      1          6        0   122
#>  [84,]  50  14      1         13        0   170
#>  [85,]  53   8      1          7        0   160
#>  [86,]  48   3      1          2        0   150
#>  [87,]  52  14      1          7        1   200
#>  [88,]  48   6      0        180        0   160
#>  [89,]  48  11      1         10        0   120
#>  [90,]  53   4      1          0        1   156
#>  [91,]  51  13      0         99        1   160
#>  [92,]  49  16      0         16        0   125
#>  [93,]  54  23      1         10        0   131
#>  [94,]  52   7      1          2        0   154
#>  [95,]  54   9      1          1        0   130
#>  [96,]  55   4      1          2        0   150
#>  [97,]  52   4      0        180        1   180
#>  [98,]  51  13      1         11        0   145
#>  [99,]  50   5      1          4        1   150
#> [100,]  52   4      0        180        0   183
#> [101,]  50   3      0        174        1   153
#> [102,]  49   6      1          0        1   130
#> [103,]  49   1      0          1        1   110
#> [104,]  50   7      1          1        0   156
#> [105,]  53   9      0          9        1    95
#> [106,]  53   8      1          0        1   130
#> [107,]  50   7      1          0        1   127
#> [108,]  55   1      0        180        0   127
#> [109,]  55   2      0          2        0   145
#> [110,]  54   1      0        180        0   162
#> [111,]  54   7      1          0        1   100
#> [112,]  56   3      0        180        1   193
#> [113,]  56   2      0        180        0   132
#> [114,]  53  18      1          9        1   150
#> [115,]  54   3      0        180        1   180
#> [116,]  52  16      0         16        0   152
#> [117,]  53  10      1          9        0   172
#> [118,]  53  15      0         15        1    90
#> [119,]  53   4      0        180        1   150
#> [120,]  55   6      0        180        1   100
#> [121,]  55   6      1          5        1   138
#> [122,]  54  12      1          0        1   190
#> [123,]  55   2      0        134        1   140
#> [124,]  54   3      0        180        0   128
#> [125,]  56   3      0          8        1   139
#> [126,]  55   1      0          2        0   130
#> [127,]  57   3      0          3        0   120
#> [128,]  54   7      1          2        0   129
#> [129,]  54   2      1          1        0   135
#> [130,]  52   9      1          3        0   170
#> [131,]  54   2      1          1        1   176
#> [132,]  57   5      1          3        1   138
#> [133,]  57   1      0          1        1   100
#> [134,]  56   4      1          0        1   140
#> [135,]  55  11      1          7        0   104
#> [136,]  52  15      1         14        0   130
#> [137,]  56  14      1         11        0   130
#> [138,]  57  10      0        180        1   170
#> [139,]  54   5      0        180        1   108
#> [140,]  55   3      1          1        1   156
#> [141,]  57   0      0          0        1   150
#> [142,]  59   3      1          1        0   172
#> [143,]  57   4      0        180        1   119
#> [144,]  53  15      1         10        1   130
#> [145,]  54  17      1          8        1   227
#> [146,]  55  13      0        166        1   140
#> [147,]  56   5      0          5        1   150
#> [148,]  54  23      1          8        0   120
#> [149,]  57   4      1          2        1   185
#> [150,]  53   4      0        147        1   145
#> [151,]  55   3      1          2        0   140
#> [152,]  55   5      0          5        1   131
#> [153,]  54   7      1          0        1   141
#> [154,]  56   4      0          4        0   164
#> [155,]  59  15      1         10        0   140
#> [156,]  58   1      1          1        1   200
#> [157,]  55   5      1          0        0   140
#> [158,]  56   7      1          5        1   120
#> [159,]  59   9      1          1        1   125
#> [160,]  57   1      0        180        0   148
#> [161,]  60  11      1          9        0   106
#> [162,]  59   3      0        180        0   120
#> [163,]  58   4      1          0        1   160
#> [164,]  57   2      0          2        1   120
#> [165,]  57   5      0        180        1   130
#> [166,]  55   5      1          0        1   160
#> [167,]  59   5      0        180        1   155
#> [168,]  59   4      1          0        1   152
#> [169,]  61   9      0          9        1   160
#> [170,]  60   0      1          0        1    80
#> [171,]  59   2      1          1        0   140
#> [172,]  58   8      0        161        1   140
#> [173,]  61   9      1          8        0   150
#> [174,]  61   3      1          2        1   102
#> [175,]  61  20      1         13        0   130
#> [176,]  57  13      1         10        0   110
#> [177,]  57   4      1          3        0   138
#> [178,]  56  14      0         45        0   130
#> [179,]  57   3      1          2        0   120
#> [180,]  58  19      1         13        1   140
#> [181,]  56  18      1         11        1   165
#> [182,]  59   9      1          0        1    80
#> [183,]  55   4      1          3        1   160
#> [184,]  56   8      1          8        0   120
#> [185,]  59  11      1          8        1   190
#> [186,]  57   1      0          1        0   126
#> [187,]  57  15      1         13        1   110
#> [188,]  58   5      1          1        1   135
#> [189,]  59  10      0        180        0   160
#> [190,]  61   8      0         77        0   120
#> [191,]  58   8      1          5        0   152
#> [192,]  62   7      1          2        1   180
#> [193,]  57   3      1          0        0   100
#> [194,]  61  18      0        170        0   140
#> [195,]  61  28      1          7        0   133
#> [196,]  58   8      1          3        1   150
#> [197,]  57   7      0        169        0   180
#> [198,]  61   7      0          7        1   150
#> [199,]  60   7      0          7        0   147
#> [200,]  61   6      0          6        0   134
#> [201,]  59  13      1          2        0   198
#> [202,]  57  12      1          9        1   120
#> [203,]  62   4      1          0        0   160
#> [204,]  60  17      1          8        1   140
#> [205,]  58   3      1          0        1   146
#> [206,]  62   4      1          3        0   173
#> [207,]  58   2      0         30        0   202
#> [208,]  63   6      0         28        1   120
#> [209,]  61   5      0          5        1   110
#> [210,]  57  18      1          9        1    93
#> [211,]  58  11      1          9        0   179
#> [212,]  57   2      1          1        0   159
#> [213,]  62  17      1         10        1   180
#> [214,]  58   7      0        180        1   150
#> [215,]  63   3      1          1        0   180
#> [216,]  63   1      0        180        1   130
#> [217,]  61   7      0        180        0   135
#> [218,]  63   4      1          3        0   222
#> [219,]  63   4      1          1        0   155
#> [220,]  60  18      1         13        0   132
#> [221,]  59   8      0        180        1   140
#> [222,]  59   1      0         22        1   162
#> [223,]  58   2      0        180        0   127
#> [224,]  59   4      0        180        0   196
#> [225,]  60   7      1          5        1   141
#> [226,]  60   7      0          7        0   140
#> [227,]  59   5      1          1        0   148
#> [228,]  60   7      1          1        1    90
#> [229,]  63   1      0          1        0   162
#> [230,]  61  15      1         13        0   170
#> [231,]  59   4      0          4        0   149
#> [232,]  64  10      1          9        0   160
#> [233,]  62   6      0          6        0   120
#> [234,]  63  12      1         10        0   200
#> [235,]  60   8      0         17        1   130
#> [236,]  61   6      1          1        1   117
#> [237,]  66   1      1          0        1   120
#> [238,]  64   6      1          0        1   140
#> [239,]  63  10      1          0        1   148
#> [240,]  65  36      1         11        0   140
#> [241,]  63   4      1          3        0   162
#> [242,]  61  10      1          2        1   194
#> [243,]  63  12      1          9        0   114
#> [244,]  63   7      0        180        0   120
#> [245,]  66   5      1          0        1   110
#> [246,]  64   0      0          0        1    90
#> [247,]  64   9      0        180        0   150
#> [248,]  65   3      0        180        1   190
#> [249,]  64   7      0        180        1   120
#> [250,]  66   6      1          1        1   130
#> [251,]  65   3      1          0        1    80
#> [252,]  63   5      1          4        0   170
#> [253,]  63   2      1          1        0   180
#> [254,]  66  16      1         11        1   169
#> [255,]  61  14      1          5        0   140
#> [256,]  61  15      1         10        0   130
#> [257,]  63   9      1          8        1   160
#> [258,]  63   3      1          2        0   120
#> [259,]  64  19      1          8        1   160
#> [260,]  67   6      0        180        1   170
#> [261,]  68   5      1          4        1   150
#> [262,]  64   6      1          0        1   125
#> [263,]  66   7      1          0        1   115
#> [264,]  64   0      0          0        1   148
#> [265,]  67   4      1          3        0   130
#> [266,]  66   6      1          0        1   140
#> [267,]  65   2      1          1        1   170
#> [268,]  68   1      0        180        1   166
#> [269,]  64  10      1          9        1   110
#> [270,]  67   8      1          1        1   130
#> [271,]  68   5      0          5        1    90
#> [272,]  63  10      0         16        1   160
#> [273,]  66  14      0        180        0   130
#> [274,]  68  18      0        180        1   260
#> [275,]  65  17      1         14        1   100
#> [276,]  63   8      1          1        1   162
#> [277,]  65  18      1          3        0   120
#> [278,]  63  10      0         18        1   130
#> [279,]  67  11      0         11        0   150
#> [280,]  68  11      0        180        0   160
#> [281,]  66  11      1          0        0   100
#> [282,]  65   4      1          2        1   145
#> [283,]  66  15      1         13        1   160
#> [284,]  69  21      1         10        0   180
#> [285,]  66   9      1          8        0   130
#> [286,]  63   8      0        180        1   120
#> [287,]  68  14      1         13        1   140
#> [288,]  68  10      1         10        1   150
#> [289,]  65   1      1          0        0   133
#> [290,]  67   7      1          4        1   130
#> [291,]  67   2      0        180        0   184
#> [292,]  65   6      0          6        0    80
#> [293,]  65  10      1          1        1   148
#> [294,]  67  12      1         12        0   160
#> [295,]  69   6      0         99        1   140
#> [296,]  64   4      0        179        0   160
#> [297,]  66   4      0        180        1   130
#> [298,]  64  11      0         11        0   125
#> [299,]  64   0      1          0        1   118
#> [300,]  67   2      0         18        0   131
#> [301,]  66   4      0        180        0   177
#> [302,]  68   4      1          0        1   160
#> [303,]  69   4      1          3        1   150
#> [304,]  65  13      1         12        1   130
#> [305,]  69  17      1         10        0   140
#> [306,]  64  21      0         21        1   155
#> [307,]  65   1      0          1        1   120
#> [308,]  68   4      0          4        1   190
#> [309,]  71   3      0          5        0   112
#> [310,]  70   7      1          0        1   190
#> [311,]  68   7      0        150        0   210
#> [312,]  71  20      1          0        1   160
#> [313,]  67   2      0        180        0   128
#> [314,]  66   9      1          3        1   151
#> [315,]  66   1      1          1        1   165
#> [316,]  70   4      1          0        1   180
#> [317,]  69   8      0        180        1   153
#> [318,]  70  14      0        171        0   166
#> [319,]  66   4      0        180        0   130
#> [320,]  67  10      1          9        0   200
#> [321,]  67   6      1          4        0   130
#> [322,]  69   0      0          0        1   148
#> [323,]  65   2      0        180        0   130
#> [324,]  67  14      1         13        0   130
#> [325,]  65  14      1         13        1   150
#> [326,]  71   7      0          7        0   230
#> [327,]  69   5      0          5        1   142
#> [328,]  69   3      0          3        1   130
#> [329,]  70  22      1         13        0   103
#> [330,]  67   1      0         36        1   104
#> [331,]  69   8      1          5        1   195
#> [332,]  69   6      1          4        1   174
#> [333,]  72   3      1          0        1   132
#> [334,]  72   7      0          7        1   110
#> [335,]  69   8      1          7        1   108
#> [336,]  67   3      0        180        0   110
#> [337,]  66   2      1          1        0   123
#> [338,]  69  19      0        180        0   130
#> [339,]  68  18      0         18        1   100
#> [340,]  67  14      0        172        1   140
#> [341,]  66   2      0        180        0   130
#> [342,]  69   8      1          2        0   121
#> [343,]  67  13      1          9        0   130
#> [344,]  70   3      0        123        0   130
#> [345,]  70   9      0        180        1   142
#> [346,]  67  22      1          1        1   140
#> [347,]  68   3      0         19        0   135
#> [348,]  67  12      1          8        0   120
#> [349,]  69   1      0          1        1   110
#> [350,]  67   1      0          1        1    60
#> [351,]  69   5      0         76        0   120
#> [352,]  72  13      1         11        1   195
#> [353,]  68  10      1          8        1   160
#> [354,]  66  24      1         13        0   130
#> [355,]  72  30      1          0        1   145
#> [356,]  68   7      1          2        0   135
#> [357,]  73  20      1          0        1   170
#> [358,]  71   6      0          9        0   120
#> [359,]  69  10      1          6        1   120
#> [360,]  70  11      0        180        1   210
#> [361,]  72  12      1         10        0   170
#> [362,]  67   8      0        180        1   170
#> [363,]  67   9      0        180        0   158
#> [364,]  73  13      0        152        1   130
#> [365,]  70   5      0        180        0   150
#> [366,]  72   2      0          2        1   100
#> [367,]  67   4      1          1        0   134
#> [368,]  72   6      1          5        0   115
#> [369,]  71   1      0        173        1   188
#> [370,]  68  23      0        180        1   220
#> [371,]  70   3      0        180        0   121
#> [372,]  71   3      1          2        0   150
#> [373,]  68   4      1          3        0   210
#> [374,]  72   5      0         28        0   120
#> [375,]  71   5      0        180        0   191
#> [376,]  69  16      1         10        1   140
#> [377,]  69   8      1          1        0   164
#> [378,]  68   7      0        180        1   130
#> [379,]  73   6      1          0        1   270
#> [380,]  72   8      1          1        1   150
#> [381,]  68  15      1         13        1   130
#> [382,]  70   3      0          3        1   159
#> [383,]  70  13      1          9        0   100
#> [384,]  73   0      0        180        1   161
#> [385,]  73   4      0        180        1   154
#> [386,]  69   2      1          0        1   110
#> [387,]  71   3      1          1        0   150
#> [388,]  71  15      1         11        0   165
#> [389,]  74  20      0         20        1   180
#> [390,]  68   9      0        180        1   120
#> [391,]  71  20      1         10        0   140
#> [392,]  73   3      1          0        1   136
#> [393,]  70   5      1          0        1   190
#> [394,]  71   3      1          2        1   190
#> [395,]  73  10      1          8        0   106
#> [396,]  69  12      1          1        1   149
#> [397,]  74   4      0          4        0   120
#> [398,]  73   4      0         58        1   160
#> [399,]  70   3      0        180        1   154
#> [400,]  73   6      0        180        0   110
#> [401,]  72  15      1          0        1   150
#> [402,]  71   7      1          2        0   143
#> [403,]  71  13      1          8        0   121
#> [404,]  70   4      1          0        1   140
#> [405,]  74   7      1          0        1   117
#> [406,]  72  10      1          8        1   153
#> [407,]  69   7      0        180        1   144
#> [408,]  71  10      1          9        1   120
#> [409,]  75   1      0          1        0   133
#> [410,]  73  10      1          9        1   146
#> [411,]  72  10      1          9        1   160
#> [412,]  74  15      1          9        1   179
#> [413,]  75   9      1          7        0   140
#> [414,]  75  13      1          1        1   130
#> [415,]  71   4      0          4        0   134
#> [416,]  72  15      1         12        1   120
#> [417,]  70   7      1          4        0   184
#> [418,]  73  10      0        180        0   162
#> [419,]  72  11      0         11        1   140
#> [420,]  70   3      0          3        0   150
#> [421,]  73  12      1         12        1   140
#> [422,]  72   2      0        180        0   120
#> [423,]  75   1      0        180        1   140
#> [424,]  72   4      1          0        1   197
#> [425,]  73   4      0        180        0   124
#> [426,]  74  34      1          8        1   233
#> [427,]  76   3      1          0        1   120
#> [428,]  71  32      1         12        1   107
#> [429,]  77  11      0         11        1   150
#> [430,]  77   4      0          4        0   185
#> [431,]  75   3      1          1        0   180
#> [432,]  72   7      1          2        0   142
#> [433,]  73  15      0         15        1   160
#> [434,]  71  16      0        180        0   140
#> [435,]  73  10      1         10        0   124
#> [436,]  74   7      0        180        1   150
#> [437,]  74   3      0          3        1   128
#> [438,]  76   1      0        180        0   114
#> [439,]  76   8      1          0        1   141
#> [440,]  74  19      1          4        1   200
#> [441,]  73   6      0          6        1   114
#> [442,]  72   4      1          3        0   160
#> [443,]  73   4      1          3        1   125
#> [444,]  76  13      1         10        0   110
#> [445,]  75   4      1          0        1   122
#> [446,]  75   7      0          7        0   190
#> [447,]  75   0      0          0        1   130
#> [448,]  73  13      1         11        0   195
#> [449,]  76  13      1          8        1   148
#> [450,]  75   4      1          2        1   188
#> [451,]  74   6      0        180        0   160
#> [452,]  75   1      0          1        1   125
#> [453,]  74   2      0        180        0   111
#> [454,]  73   1      0         52        1   105
#> [455,]  73   0      0        180        0   156
#> [456,]  78  12      1         11        1   160
#> [457,]  76  44      1         10        0   105
#> [458,]  76   5      0        180        0   185
#> [459,]  74  10      1          0        1   135
#> [460,]  74   8      1          8        1   170
#> [461,]  75   9      0        180        1   140
#> [462,]  73  33      1         12        1   175
#> [463,]  73  10      1          9        0   146
#> [464,]  77  12      0        180        0   130
#> [465,]  76  12      1         11        1   120
#> [466,]  73   7      1          0        0   174
#> [467,]  74   6      0         79        1   140
#> [468,]  75   3      1          1        1   171
#> [469,]  74   9      1          8        0   200
#> [470,]  75   6      0        180        0   150
#> [471,]  79  10      1          8        0   190
#> [472,]  74   2      1          0        1   130
#> [473,]  78  18      0         18        1   144
#> [474,]  76  29      0         47        0    90
#> [475,]  73  11      1          2        1   110
#> [476,]  74   2      0        180        0   100
#> [477,]  78   7      0          7        1   133
#> [478,]  74  15      0        180        1   172
#> [479,]  78   8      1          6        1   110
#> [480,]  76  13      1          1        1   170
#> [481,]  78  32      1          9        1   198
#> [482,]  79   6      0        180        0   170
#> [483,]  80  10      1          6        1   147
#> [484,]  78   0      0        180        1   212
#> [485,]  78  13      1          5        0   130
#> [486,]  75   5      0        119        1   150
#> [487,]  75  12      1          1        1   120
#> [488,]  78  15      0        180        1   270
#> [489,]  75  13      1          6        0   150
#> [490,]  74  10      1          8        0   135
#> [491,]  76   1      0          1        1    83
#> [492,]  75   4      1          0        0   212
#> [493,]  75  11      1          4        0   162
#> [494,]  76   7      0         29        1   150
#> [495,]  79   8      0         32        1   120
#> [496,]  80   6      0          6        1   150
#> [497,]  76   3      1          0        1   140
#> [498,]  78  14      1          0        1   140
#> [499,]  81   1      0          1        0   130
#> [500,]  79   4      0          4        1   125
#> [501,]  76  12      1         10        1   127
#> [502,]  80   3      1          0        1   120
#> [503,]  78  11      0        180        1   135
#> [504,]  76   1      0          1        1   140
#> [505,]  77  31      1          3        1   161
#> [506,]  78   7      1          0        1   110
#> [507,]  77   7      0        180        1   170
#> [508,]  77   6      0          6        1   144
#> [509,]  79   4      1          0        1   120
#> [510,]  80  15      1         12        1   150
#> [511,]  77   9      1          4        0   141
#> [512,]  82   5      0          8        1   120
#> [513,]  80  40      1          0        1   138
#> [514,]  80  17      1         12        0   100
#> [515,]  79  10      0         10        1   120
#> [516,]  80  15      1          0        1    90
#> [517,]  81   4      1          2        1   126
#> [518,]  78  32      0        180        1   130
#> [519,]  79   1      0         37        1   140
#> [520,]  81   3      0        180        0   184
#> [521,]  81   2      0        175        0   172
#> [522,]  78   7      0          7        1   147
#> [523,]  77  13      1          0        1   190
#> [524,]  78  15      0         15        0   165
#> [525,]  78   4      0        180        0   175
#> [526,]  79   3      0          3        1   101
#> [527,]  78  26      1          5        0   194
#> [528,]  80   1      0          1        0   100
#> [529,]  82   3      1          1        1   144
#> [530,]  77   5      0         85        0   188
#> [531,]  80   2      1          1        0   168
#> [532,]  80   6      1          0        1   119
#> [533,]  80   5      0          5        1   130
#> [534,]  82   1      1          0        1    82
#> [535,]  79  10      0        180        1   150
#> [536,]  77   4      0        180        1    98
#> [537,]  81   1      0        108        0   129
#> [538,]  78  12      0        180        0   134
#> [539,]  79   1      0        125        0   193
#> [540,]  82  21      1          2        0   155
#> [541,]  79   4      0          4        1   121
#> [542,]  80   6      0          6        1   110
#> [543,]  83   5      0        180        0   148
#> [544,]  81  11      1          8        0   160
#> [545,]  81   5      0        177        0    41
#> [546,]  78  23      1         10        1   145
#> [547,]  78   9      1          4        1   120
#> [548,]  79   1      0        180        1   170
#> [549,]  81  15      0        180        1   140
#> [550,]  80   7      1          0        1   146
#> [551,]  84   5      1          1        1    85
#> [552,]  81  20      1          9        0   170
#> [553,]  81  16      0         16        1   110
#> [554,]  80   6      1          0        1   150
#> [555,]  81   8      0        180        0   146
#> [556,]  80   8      1          7        0   160
#> [557,]  79   7      0        177        0   197
#> [558,]  79   0      1          0        1    96
#> [559,]  85   4      0        180        0    90
#> [560,]  84   4      0        167        0   198
#> [561,]  80   3      1          1        1   230
#> [562,]  82  23      1          0        0   110
#> [563,]  83   3      0        180        0   174
#> [564,]  85   3      1          2        1   160
#> [565,]  80  13      1          8        1   140
#> [566,]  84   4      0         89        1   129
#> [567,]  80   2      1          0        1   130
#> [568,]  79   4      0          4        1    60
#> [569,]  80   6      0         71        1   189
#> [570,]  83   1      0          1        1   100
#> [571,]  82  19      0         19        0   120
#> [572,]  80  30      1         13        0   220
#> [573,]  83   9      0        180        0   198
#> [574,]  79  14      1          0        0   110
#> [575,]  83   2      0        154        0   130
#> [576,]  82   0      0          2        1   100
#> [577,]  85   9      1          6        1   160
#> [578,]  83   1      0        180        0   160
#> [579,]  81   4      0          4        0   175
#> [580,]  84  15      1         13        1   110
#> [581,]  81   1      0          1        1   145
#> [582,]  81  12      0         12        1   163
#> [583,]  82   5      1          0        1   146
#> [584,]  81   4      0          4        0   160
#> [585,]  86  12      0        180        1   120
#> [586,]  83  12      1          2        1   170
#> [587,]  82  15      1          0        0   183
#> [588,]  80   2      0         88        0   135
#> [589,]  86   8      0          8        1   132
#> [590,]  81  16      1          9        0   180
#> [591,]  84   6      0        165        0   145
#> [592,]  86   3      0          3        1   140
#> [593,]  82   9      0        180        1   134
#> [594,]  84   3      0        180        1   120
#> [595,]  83   9      0        180        1   149
#> [596,]  82   1      0        180        1   193
#> [597,]  82  14      1         11        1   103
#> [598,]  86   6      1          0        1   140
#> [599,]  83  19      0         43        0   150
#> [600,]  84   3      1          2        0   125
#> [601,]  83  10      1          0        1   190
#> [602,]  86   2      0        180        1   169
#> [603,]  88  14      1          3        1   130
#> [604,]  84   3      0          3        1   121
#> [605,]  83  13      1         12        0   170
#> [606,]  84   7      1          2        0   148
#> [607,]  87   2      0        180        0   113
#> [608,]  84   9      0         92        1   110
#> [609,]  84   3      0        180        1   170
#> [610,]  86   4      0         38        1   122
#> [611,]  86  13      0        177        0   163
#> [612,]  85   3      0          3        1   113
#> [613,]  84  13      0         62        1   100
#> [614,]  86   6      1          1        0   112
#> [615,]  88   4      0          4        0   100
#> [616,]  83  20      1          3        1   150
#> [617,]  88   4      0          4        1   115
#> [618,]  85  22      0         22        1   184
#> [619,]  83   9      0         65        1   150
#> [620,]  86   9      1          7        1   142
#> [621,]  87   2      0        180        1   130
#> [622,]  86   6      0         46        0   173
#> [623,]  88   3      0        115        0   110
#> [624,]  88   2      0        180        1    68
#> [625,]  83   3      0          3        1   130
#> [626,]  87   8      0          8        1   157
#> [627,]  88   4      0          4        0    86
#> [628,]  89   4      0          4        1   153
#> [629,]  89   5      0        119        1   140
#> [630,]  87   1      0          1        0   170
#> [631,]  84   8      0        180        1   119
#> [632,]  85   8      0          8        1   136
#> [633,]  87  15      1          9        1   138
#> [634,]  84   0      0        180        1   136
#> [635,]  89  10      0         46        1   170
#> [636,]  90  14      0         14        1   100
#> [637,]  88   1      0          1        0   135
#> [638,]  86   4      0        180        1   145
#> [639,]  91   8      0          8        0   100
#> [640,]  86   3      1          0        1    80
#> [641,]  88   8      0         50        1   154
#> [642,]  90  11      1         10        1   186
#> [643,]  86   9      1          7        0   130
#> [644,]  90   4      1          0        0   121
#> [645,]  91   1      0          1        1    74
#> [646,]  87  43      0        178        1   130
#> [647,]  87   5      0         36        1   150
#> [648,]  90   5      1          0        1   125
#> [649,]  88   3      1          2        0   159
#> [650,]  89   3      1          1        1   160
#> [651,]  91   3      0         33        1   137
#> [652,]  88   5      0        158        0   100
#> [653,]  91   5      0        169        1   176
#> [654,]  92   7      0          7        1   110
#> [655,]  89   4      0          4        1   159
#> [656,]  91   0      0          0        0     0
#> [657,]  90  18      0        180        0   188
#> [658,]  94   6      0         50        0    78
#> [659,]  90   1      0          1        1   118
#> [660,]  94   8      0          8        1   142
#> [661,]  91   1      0        180        0   158
#> [662,]  90  16      0         16        1   106
#> [663,]  96   3      0         12        1    97
#> [664,]  95   8      1          5        1   150
#> [665,]  94   3      0         26        1   144
#> [666,]  91  12      0         53        1   212
#> [667,]  92   5      0         69        0   139
#> [668,]  92   2      0          2        0   112
#> [669,]  93   4      0        180        1   135
#> [670,]  96  15      1          0        1   140
#> 
#> $y
#>   [1] 180.0+   5.0+   5.0+   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  12.0 
#>  [11]   5.0+ 180.0+ 180.0+ 180.0+   2.0+ 180.0+ 180.0+ 180.0+   3.0  180.0+
#>  [21] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 180.0+ 155.0+ 180.0+ 180.0+
#>  [31]   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 150.0  180.0+ 180.0+   6.0+ 180.0+
#>  [41] 161.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 177.0+ 180.0+ 180.0+
#>  [51] 180.0+ 180.0+ 180.0+  10.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [61]   2.0  179.0+ 179.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+ 180.0+ 180.0+
#>  [71] 180.0+ 180.0+ 180.0+ 180.0+  88.0+ 180.0+   4.0+ 180.0+ 180.0+ 180.0+
#>  [81] 180.0+   5.0  180.0+ 180.0+ 180.0+ 180.0+  85.0  180.0+ 180.0+ 166.0+
#>  [91]  99.0   16.0+ 152.0+   7.0+ 180.0+ 180.0+ 180.0+  13.0+ 171.0+ 180.0+
#> [101] 174.0+   6.0+   1.0  180.0+   9.0+ 180.0+ 180.0+ 180.0+   2.0  180.0+
#> [111]   7.0+ 180.0+ 180.0+ 180.0+ 180.0+  16.0+ 180.0+  15.0+ 180.0+ 180.0+
#> [121] 180.0+  12.0+ 134.0+ 180.0+   8.0    2.0    3.0+ 180.0+ 180.0+ 180.0+
#> [131] 180.0+ 140.0    1.0  165.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [141]   0.5  180.0+ 180.0+ 180.0+ 171.0+ 166.0+   5.0+ 180.0+   4.0+ 147.0+
#> [151] 180.0+   5.0+ 180.0+   4.0+ 180.0+   1.0  180.0+ 180.0+ 180.0+ 180.0+
#> [161] 180.0+ 180.0+ 180.0+   2.0  180.0+ 180.0+ 180.0+ 180.0+   9.0+   0.5 
#> [171] 180.0+ 161.0+ 180.0+   3.0  180.0+ 180.0+ 180.0+  45.0    3.0+  19.0 
#> [181] 180.0+   9.0+ 180.0+   8.0  180.0+   1.0+  15.0  180.0+ 180.0+  77.0 
#> [191]   8.0+ 180.0+ 180.0+ 170.0   94.0  180.0+ 169.0    7.0    7.0+   6.0 
#> [201] 180.0+ 180.0+ 180.0+ 180.0+   3.0+ 180.0+  30.0   28.0    5.0   18.0 
#> [211] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+
#> [221] 180.0+  22.0  180.0+ 180.0+  84.0    7.0+ 180.0+ 180.0+   1.0  180.0+
#> [231]   4.0+ 167.0    6.0+ 180.0+  17.0  180.0+ 180.0+ 180.0+ 180.0+  36.0 
#> [241] 180.0+  88.0   12.0  180.0+ 180.0+   0.5  180.0+ 180.0+ 180.0+ 180.0+
#> [251]   3.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   3.0+ 103.0  180.0+
#> [261]   5.0+ 180.0+ 179.0+   0.5+ 180.0+ 180.0+ 175.0+ 180.0+ 180.0+   8.0 
#> [271]   5.0   16.0  180.0+ 180.0+ 180.0+ 180.0+ 123.0+  18.0   11.0+ 180.0+
#> [281] 180.0+   4.0+ 180.0+ 174.0+ 180.0+ 180.0+ 180.0+  10.0  180.0+ 180.0+
#> [291] 180.0+   6.0  180.0+  12.0   99.0  179.0+ 180.0+  11.0+   0.5   18.0 
#> [301] 180.0+ 180.0+ 152.0+ 180.0+ 180.0+  21.0+   1.0    4.0    5.0    7.0+
#> [311] 150.0  180.0+ 180.0+ 180.0+   1.0  180.0+ 180.0+ 171.0  180.0+ 174.0+
#> [321]   6.0    0.5  180.0+ 180.0+  14.0+   7.0+   5.0+   3.0+ 180.0+  36.0 
#> [331] 180.0+  97.0  180.0+   7.0    8.0+ 180.0+   2.0+ 180.0+  18.0  172.0+
#> [341] 180.0+   8.0+  13.0+ 123.0  180.0+  51.0   19.0  180.0+   1.0    1.0 
#> [351]  76.0  132.0   10.0+ 180.0+ 162.0    7.0+ 124.0    9.0  180.0+ 180.0+
#> [361]  12.0  180.0+ 180.0+ 152.0  180.0+   2.0   76.0  180.0+ 173.0+ 180.0+
#> [371] 180.0+ 180.0+ 180.0+  28.0  180.0+  16.0+ 180.0+ 180.0+   6.0  180.0+
#> [381]  15.0    3.0+  13.0+ 180.0+ 180.0+   2.0    3.0+ 180.0+  20.0  180.0+
#> [391]  20.0  180.0+ 180.0+   3.0   87.0   12.0    4.0+  58.0  180.0+ 180.0+
#> [401] 180.0+ 180.0+ 175.0  180.0+ 180.0+  10.0+ 180.0+ 179.0+   1.0  180.0+
#> [411] 159.0  180.0+ 180.0+  13.0    4.0+ 180.0+ 104.0+ 180.0+  11.0    3.0+
#> [421]  12.0  180.0+ 180.0+ 180.0+ 180.0+  34.0  180.0+ 177.0+  11.0+   4.0+
#> [431] 180.0+   7.0   15.0+ 180.0+  10.0  180.0+   3.0  180.0+ 180.0+ 180.0+
#> [441]   6.0  180.0+ 180.0+ 174.0+   4.0    7.0    0.5  180.0+ 180.0+  46.0 
#> [451] 180.0+   1.0  180.0+  52.0  180.0+  12.0  180.0+ 180.0+ 180.0+   8.0 
#> [461] 180.0+  33.0  180.0+ 180.0+  12.0    7.0+  79.0    3.0  168.0+ 180.0+
#> [471] 180.0+ 176.0+  18.0   47.0   11.0  180.0+   7.0  180.0+   8.0+ 180.0+
#> [481]  32.0  180.0+  10.0  180.0+ 172.0  119.0   12.0  180.0+ 180.0+ 180.0+
#> [491]   1.0    4.0+ 152.0+  29.0   32.0    6.0    3.0+ 180.0+   1.0    4.0 
#> [501] 180.0+   3.0+ 180.0+   1.0  171.0   43.0  180.0+   6.0  138.0  180.0+
#> [511]  71.0    8.0   40.0   17.0   10.0+ 180.0+  93.0  180.0+  37.0  180.0+
#> [521] 175.0+   7.0+  22.0   15.0+ 180.0+   3.0  171.0+   1.0  180.0+  85.0 
#> [531]  10.0    6.0    5.0    1.0  180.0+ 180.0+ 108.0  180.0+ 125.0  180.0+
#> [541]   4.0    6.0  180.0+ 180.0+ 177.0+  70.0  180.0+ 180.0+ 180.0+   7.0+
#> [551] 180.0+  20.0   16.0  180.0+ 180.0+ 180.0+ 177.0+   0.5  180.0+ 167.0 
#> [561]   3.0+  62.0  180.0+ 180.0+ 180.0+  89.0  180.0+   4.0   71.0    1.0 
#> [571]  19.0   30.0  180.0+ 180.0+ 154.0    2.0  180.0+ 180.0+   4.0+ 180.0+
#> [581]   1.0   12.0    5.0+   4.0+ 180.0+  77.0   83.0   88.0    8.0  180.0+
#> [591] 165.0    3.0  180.0+ 180.0+ 180.0+ 180.0+ 174.0    6.0   43.0  180.0+
#> [601] 180.0+ 180.0+  14.0    3.0   13.0  180.0+ 180.0+  92.0  180.0+  38.0 
#> [611] 177.0    3.0+  62.0    6.0+   4.0+  20.0    4.0   22.0   65.0   11.0 
#> [621] 180.0+  46.0  115.0  180.0+   3.0+   8.0+   4.0    4.0  119.0    1.0+
#> [631] 180.0+   8.0  180.0+ 180.0+  46.0   14.0    1.0+ 180.0+   8.0    3.0 
#> [641]  50.0   11.0  180.0+   4.0    1.0  178.0+  36.0   89.0   75.0    3.0+
#> [651]  33.0  158.0  169.0    7.0    4.0    0.5  180.0+  50.0    1.0+   8.0+
#> [661] 180.0+  16.0   12.0    8.0   26.0   53.0   69.0    2.0  180.0+  15.0+
#> 
#> $weights
#> NULL
#> 
#> $offset
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8219866