Cross-Validated GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::cv.glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.cv.glmnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.cv.glmnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.cv_glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

foldid	untyped	NULL		-

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nfolds	integer	10		\([3, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	TRUE	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	lambda.1se		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

standardize.response	logical	FALSE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.gaussian	character	-	covariance, naive	-

type.logistic	character	Newton	Newton, modified.Newton	-

type.measure	character	deviance	deviance, C	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.coxnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvCVGlmnet

Methods

Public methods

LearnerSurvCVGlmnet$new()

LearnerSurvCVGlmnet$selected_features()

LearnerSurvCVGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvCVGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvCVGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvCVGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("surv.cv_glmnet")
print(learner)
#> 
#> ── <LearnerSurvCVGlmnet> (surv.cv_glmnet): Regularized Generalized Linear Model 
#> • Model: -
#> • Parameters: use_pred_offset=TRUE
#> • Packages: mlr3, mlr3proba, mlr3extralearners, and glmnet
#> • Predict Types: [crank], distr, and lp
#> • Feature Types: logical, integer, and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: offset, selected_features, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("grace")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#> Measure: Partial Likelihood Deviance 
#> 
#>      Lambda Index Measure      SE Nonzero
#> min 0.00309    44   2.562 0.09715       6
#> 1se 0.05530    13   2.652 0.09356       4
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  32   5      1          0        1   121
#>   [2,]  35   5      1          2        0   172
#>   [3,]  35  10      1          9        0   106
#>   [4,]  34   5      0          5        0   120
#>   [5,]  35   2      0        180        0   121
#>   [6,]  35   2      1          1        1   112
#>   [7,]  36   1      0        180        1   155
#>   [8,]  35   0      0        180        1   119
#>   [9,]  38  12      1          8        1   120
#>  [10,]  36   5      1          0        1   115
#>  [11,]  38  16      1         10        0   160
#>  [12,]  38  12      1         11        1    92
#>  [13,]  42   3      1          1        1   130
#>  [14,]  37   1      1          0        1   146
#>  [15,]  40   2      1          1        1   148
#>  [16,]  42   2      0        180        1   100
#>  [17,]  38   5      1          3        0   125
#>  [18,]  40   6      0        180        1   138
#>  [19,]  40  11      1         10        1   120
#>  [20,]  43   3      1          0        1   100
#>  [21,]  41   2      1          1        0   166
#>  [22,]  40   1      1          0        1   145
#>  [23,]  43   4      1          0        1   130
#>  [24,]  42   4      0        180        0   162
#>  [25,]  40   3      1          1        0   170
#>  [26,]  43   2      1          1        1   116
#>  [27,]  42   2      0        180        1   124
#>  [28,]  44   5      1          1        0   170
#>  [29,]  41  13      1          1        0   140
#>  [30,]  45   9      1          7        0   110
#>  [31,]  45   6      0        180        1   170
#>  [32,]  45   2      0        180        1   140
#>  [33,]  47   4      1          3        0   118
#>  [34,]  48  15      0        180        1   160
#>  [35,]  45   3      0        150        0   130
#>  [36,]  46   7      1          2        0   166
#>  [37,]  43  29      0        180        1   180
#>  [38,]  45   4      1          0        0   124
#>  [39,]  46  13      1         10        0   100
#>  [40,]  44   0      1          0        1    96
#>  [41,]  47   4      1          3        1   160
#>  [42,]  43   3      1          0        1   124
#>  [43,]  45   8      1          0        1   117
#>  [44,]  49   5      0         73        1   136
#>  [45,]  46   2      1          1        1   122
#>  [46,]  47   2      0        180        0   108
#>  [47,]  44   9      1          8        1   135
#>  [48,]  45   5      0        180        1   190
#>  [49,]  46   5      1          3        0   130
#>  [50,]  44   2      0        180        0   142
#>  [51,]  46  15      0        180        1   120
#>  [52,]  45   9      1          0        1   145
#>  [53,]  48   3      0        180        0   154
#>  [54,]  48  12      1         11        0   200
#>  [55,]  46   3      1          0        1   119
#>  [56,]  49   4      0        180        0   117
#>  [57,]  47  10      0         10        1   140
#>  [58,]  50   1      1          0        1   129
#>  [59,]  48   2      1          0        0   184
#>  [60,]  47   7      0        180        0   145
#>  [61,]  50   4      1          1        0   125
#>  [62,]  49   7      1          7        1   110
#>  [63,]  46   9      1          9        1   122
#>  [64,]  49   2      0          2        0   105
#>  [65,]  49  15      1         11        1   160
#>  [66,]  49  23      0        179        1   112
#>  [67,]  46   6      1          0        1   156
#>  [68,]  52   2      0        180        1   170
#>  [69,]  50   4      0          4        1   100
#>  [70,]  51   3      1          2        0   113
#>  [71,]  50   1      1          0        0   150
#>  [72,]  50   9      0        180        0   130
#>  [73,]  47   6      0        180        1   162
#>  [74,]  51   8      0        180        1   140
#>  [75,]  52   2      0        180        0   155
#>  [76,]  46   3      0        180        1   120
#>  [77,]  46   1      1          1        0   145
#>  [78,]  50   4      1          1        0   150
#>  [79,]  48   7      1          0        1   110
#>  [80,]  52   4      1          4        0   152
#>  [81,]  53   5      0         77        0   159
#>  [82,]  53   7      1          0        0   199
#>  [83,]  54   6      1          3        0   129
#>  [84,]  51   3      1          1        0   140
#>  [85,]  50   2      0          5        1   106
#>  [86,]  50  10      1          6        0   122
#>  [87,]  48   3      1          2        0   150
#>  [88,]  51  25      1          1        0   202
#>  [89,]  49   5      1          2        1   150
#>  [90,]  53   4      0          4        0   140
#>  [91,]  52  14      1          7        1   200
#>  [92,]  53   4      1          0        1   156
#>  [93,]  49  16      0         16        0   125
#>  [94,]  55   3      1          1        0   150
#>  [95,]  54  23      1         10        0   131
#>  [96,]  52   7      1          2        0   154
#>  [97,]  55   6      1          2        1   114
#>  [98,]  54   9      1          1        0   130
#>  [99,]  55   4      1          2        0   150
#> [100,]  52   4      0        180        1   180
#> [101,]  51  13      1         11        0   145
#> [102,]  54   4      1          0        1   121
#> [103,]  50   3      0        174        1   153
#> [104,]  49   6      1          0        1   130
#> [105,]  49   1      0          1        1   110
#> [106,]  53   9      0          9        1    95
#> [107,]  53   8      1          0        1   130
#> [108,]  50   7      1          0        1   127
#> [109,]  52   5      0        175        1   117
#> [110,]  55   1      0        180        0   127
#> [111,]  55   2      0          2        0   145
#> [112,]  54   1      0        180        0   162
#> [113,]  55   5      1          4        1   120
#> [114,]  52   8      0        180        0   119
#> [115,]  55   6      0        180        0   170
#> [116,]  52  16      0         16        0   152
#> [117,]  53  10      1          9        0   172
#> [118,]  52  16      1         14        0   170
#> [119,]  53   4      0        180        1   150
#> [120,]  55   6      0        180        1   100
#> [121,]  55   6      1          5        1   138
#> [122,]  54  12      1          0        1   190
#> [123,]  55   2      0        134        1   140
#> [124,]  54   3      0        180        0   128
#> [125,]  56   3      0          8        1   139
#> [126,]  55   1      0          2        0   130
#> [127,]  54   7      1          2        0   129
#> [128,]  52   9      1          3        0   170
#> [129,]  54   2      1          1        1   176
#> [130,]  57   1      0        180        1   156
#> [131,]  57   1      0          1        1   100
#> [132,]  52   2      0        180        0   140
#> [133,]  55  11      1          7        0   104
#> [134,]  52  15      1         14        0   130
#> [135,]  53   3      1          0        1   200
#> [136,]  57  10      0        180        1   170
#> [137,]  58   8      0          8        1   130
#> [138,]  54   5      0        180        1   108
#> [139,]  55   3      1          1        1   156
#> [140,]  57   0      0          0        1   150
#> [141,]  53  21      1         13        1   130
#> [142,]  59   3      1          1        0   172
#> [143,]  58   6      1          0        1    90
#> [144,]  53  15      1         10        1   130
#> [145,]  54  17      1          8        1   227
#> [146,]  55   9      1          2        1   147
#> [147,]  55  13      0        166        1   140
#> [148,]  56   5      0          5        1   150
#> [149,]  54  23      1          8        0   120
#> [150,]  53   4      0        147        1   145
#> [151,]  57  11      1         10        1   129
#> [152,]  55   3      1          2        0   140
#> [153,]  54   7      1          0        1   141
#> [154,]  56   4      0          4        0   164
#> [155,]  59  15      1         10        0   140
#> [156,]  58   9      1          0        1   180
#> [157,]  56   7      1          5        1   120
#> [158,]  55   2      0          2        0   106
#> [159,]  59   9      1          1        1   125
#> [160,]  57   1      0        180        0   148
#> [161,]  60  11      1          9        0   106
#> [162,]  59   3      0        180        0   120
#> [163,]  58   4      1          0        1   160
#> [164,]  60   5      1          1        0   138
#> [165,]  57   5      0        180        1   130
#> [166,]  58  11      1          9        1   124
#> [167,]  55   5      1          0        1   160
#> [168,]  57  10      1          9        0   103
#> [169,]  59   4      1          0        1   152
#> [170,]  58  26      1          0        1   189
#> [171,]  61   9      0          9        1   160
#> [172,]  58   4      1          3        0   120
#> [173,]  60   0      1          0        1    80
#> [174,]  59   2      1          1        0   140
#> [175,]  58  14      1          6        0   190
#> [176,]  61   4      1          3        0   151
#> [177,]  61   9      1          8        0   150
#> [178,]  61   3      1          2        1   102
#> [179,]  58   1      0          1        1   100
#> [180,]  57  13      1         10        0   110
#> [181,]  57   2      1          0        1   116
#> [182,]  58  10      0         10        1   150
#> [183,]  57   4      1          3        0   138
#> [184,]  57  11      0        180        1   150
#> [185,]  61   3      0         17        0   143
#> [186,]  56  14      0         45        0   130
#> [187,]  57   3      1          2        0   120
#> [188,]  58  19      1         13        1   140
#> [189,]  56  18      1         11        1   165
#> [190,]  59   9      1          0        1    80
#> [191,]  55   4      1          3        1   160
#> [192,]  58  11      0        172        1   135
#> [193,]  60  12      1          0        1   114
#> [194,]  57   1      0          1        0   126
#> [195,]  59   5      1          2        0   182
#> [196,]  58   5      1          1        1   135
#> [197,]  59  10      0        180        0   160
#> [198,]  61  13      0         13        0   210
#> [199,]  62  10      1          0        1   153
#> [200,]  62   7      1          2        1   180
#> [201,]  57   3      1          0        0   100
#> [202,]  61  18      0        170        0   140
#> [203,]  61  28      1          7        0   133
#> [204,]  61   6      0          6        0   134
#> [205,]  59  13      1          2        0   198
#> [206,]  60  17      1          8        1   140
#> [207,]  58   3      1          0        1   146
#> [208,]  62   4      1          3        0   173
#> [209,]  59  16      1          9        1   133
#> [210,]  63   6      0         28        1   120
#> [211,]  61  13      0         13        0   120
#> [212,]  61   5      0          5        1   110
#> [213,]  58  11      1          9        0   179
#> [214,]  62  17      1         10        1   180
#> [215,]  63   4      1          3        0   222
#> [216,]  64   4      0        180        0   130
#> [217,]  60  18      1         13        0   132
#> [218,]  59   8      0        180        1   140
#> [219,]  61   9      1          9        1   150
#> [220,]  58   9      1          9        0   110
#> [221,]  62   7      0          7        0   150
#> [222,]  59   1      0         22        1   162
#> [223,]  58   2      0        180        0   127
#> [224,]  59   4      0        180        0   196
#> [225,]  60   7      1          5        1   141
#> [226,]  60   7      0          7        0   140
#> [227,]  65  13      0        180        1   100
#> [228,]  63   1      0          1        0   162
#> [229,]  63   1      0          1        0   130
#> [230,]  59   4      0          4        0   149
#> [231,]  64  10      1          9        0   160
#> [232,]  62   6      0          6        0   120
#> [233,]  63  12      1         10        0   200
#> [234,]  59  10      0        180        1   130
#> [235,]  60   8      0         17        1   130
#> [236,]  61   6      1          1        1   117
#> [237,]  64  12      1         11        0   160
#> [238,]  66   1      1          0        1   120
#> [239,]  64   6      1          0        1   140
#> [240,]  63  10      1          0        1   148
#> [241,]  63  14      1          9        0   123
#> [242,]  65  36      1         11        0   140
#> [243,]  63   4      1          3        0   162
#> [244,]  66   3      1          1        0   127
#> [245,]  63  12      1          9        0   114
#> [246,]  63   7      0        180        0   120
#> [247,]  66   5      1          0        1   110
#> [248,]  65   8      1          0        0   168
#> [249,]  60   6      0        180        0   130
#> [250,]  61  12      1         11        0   154
#> [251,]  64   9      0        180        0   150
#> [252,]  61   4      0        180        1   113
#> [253,]  63  16      1          7        1   110
#> [254,]  64   7      0        180        1   120
#> [255,]  66   6      1          1        1   130
#> [256,]  62   3      1          1        1   199
#> [257,]  63   5      1          4        0   170
#> [258,]  63   2      1          1        0   180
#> [259,]  62  13      1         11        0   180
#> [260,]  64   2      0          2        0   201
#> [261,]  66  16      1         11        1   169
#> [262,]  62   9      0        180        0   145
#> [263,]  61  15      1         10        0   130
#> [264,]  63   9      1          8        1   160
#> [265,]  63   3      1          2        0   120
#> [266,]  63   2      1          0        0   140
#> [267,]  64  19      1          8        1   160
#> [268,]  68   5      1          4        1   150
#> [269,]  64  13      1         12        1   150
#> [270,]  64   6      1          0        1   125
#> [271,]  66   7      1          0        1   115
#> [272,]  66  13      1          0        0   118
#> [273,]  64  14      1         13        1   150
#> [274,]  65   3      0          3        0   105
#> [275,]  67   4      1          3        0   130
#> [276,]  66   6      1          0        1   140
#> [277,]  65   2      1          1        1   170
#> [278,]  68   1      0        180        1   166
#> [279,]  64  10      1          9        1   110
#> [280,]  63   7      1          0        0   162
#> [281,]  67   8      1          1        1   130
#> [282,]  68   5      0          5        1    90
#> [283,]  66  14      0        180        0   130
#> [284,]  64   1      0          1        1   120
#> [285,]  68  18      0        180        1   260
#> [286,]  65  17      1         14        1   100
#> [287,]  63   1      1          0        1   155
#> [288,]  63  10      0         18        1   130
#> [289,]  67  11      0         11        0   150
#> [290,]  68  11      0        180        0   160
#> [291,]  68  14      0         79        0   172
#> [292,]  65  15      1         12        1   150
#> [293,]  65   4      1          2        1   145
#> [294,]  69  12      0         15        1   140
#> [295,]  66  15      1         13        1   160
#> [296,]  63   2      0        180        0   150
#> [297,]  69   6      0        180        1   100
#> [298,]  66   9      1          8        0   130
#> [299,]  65   8      1          0        1    90
#> [300,]  66   3      0          3        1   138
#> [301,]  68  10      1         10        1   150
#> [302,]  67   7      1          4        1   130
#> [303,]  67   2      0        180        0   184
#> [304,]  65  10      1          1        1   148
#> [305,]  66  19      1         12        1   150
#> [306,]  67  12      1         12        0   160
#> [307,]  65   4      1          1        0   130
#> [308,]  64   4      0        179        0   160
#> [309,]  70  15      1         12        1   132
#> [310,]  64  11      0         11        0   125
#> [311,]  64   0      1          0        1   118
#> [312,]  66   4      0        180        0   177
#> [313,]  69   4      1          3        1   150
#> [314,]  65  13      1         12        1   130
#> [315,]  69   8      0         93        0   140
#> [316,]  64  21      0         21        1   155
#> [317,]  68  18      1          0        1   160
#> [318,]  68   4      0          4        1   190
#> [319,]  70   7      1          0        1   190
#> [320,]  68   7      0        150        0   210
#> [321,]  71  20      1          0        1   160
#> [322,]  70   4      1          0        1   180
#> [323,]  66   4      0        180        0   130
#> [324,]  67  10      1          9        0   200
#> [325,]  67   6      1          4        0   130
#> [326,]  68  18      1         14        1   170
#> [327,]  69   0      0          0        1   148
#> [328,]  65   2      0        180        0   130
#> [329,]  68   7      1          0        1   150
#> [330,]  65  14      1         13        1   150
#> [331,]  71   6      0         45        1   158
#> [332,]  71   3      0        103        0   133
#> [333,]  70  22      1         13        0   103
#> [334,]  67   1      0         36        1   104
#> [335,]  67   5      0          5        0   130
#> [336,]  68   6      0        180        0   145
#> [337,]  69   8      1          5        1   195
#> [338,]  69   6      1          4        1   174
#> [339,]  72   3      1          0        1   132
#> [340,]  72   7      0          7        1   110
#> [341,]  69   8      1          7        1   108
#> [342,]  66   2      1          1        0   123
#> [343,]  69  19      0        180        0   130
#> [344,]  67  14      0        172        1   140
#> [345,]  69  11      1          0        1   120
#> [346,]  67   7      1          4        0   122
#> [347,]  69   4      1          3        0   132
#> [348,]  69   8      1          2        0   121
#> [349,]  67  13      1          9        0   130
#> [350,]  70   3      0        123        0   130
#> [351,]  70   9      0        180        1   142
#> [352,]  72   5      1          4        0   170
#> [353,]  67  22      1          1        1   140
#> [354,]  67  12      1          8        0   120
#> [355,]  67   4      0         60        1   136
#> [356,]  69   5      0         76        0   120
#> [357,]  67   8      1          0        1   130
#> [358,]  72  13      1         11        1   195
#> [359,]  68  10      1          8        1   160
#> [360,]  70  35      1          0        1   105
#> [361,]  72  30      1          0        1   145
#> [362,]  70   7      0          7        0   102
#> [363,]  68   7      1          2        0   135
#> [364,]  71   6      0          9        0   120
#> [365,]  69  10      1          6        1   120
#> [366,]  70  11      0        180        1   210
#> [367,]  72  19      1          8        0   120
#> [368,]  72  12      1         10        0   170
#> [369,]  67   8      0        180        1   170
#> [370,]  67   5      1          0        1   147
#> [371,]  67   9      0        180        0   158
#> [372,]  70   5      0        180        0   150
#> [373,]  72   2      0          2        1   100
#> [374,]  67   4      1          1        0   134
#> [375,]  71   1      0        173        1   188
#> [376,]  68  23      0        180        1   220
#> [377,]  69   3      0        180        0   220
#> [378,]  71   3      1          2        0   150
#> [379,]  68   4      1          3        0   210
#> [380,]  72   5      0         28        0   120
#> [381,]  73   6      0        180        1   117
#> [382,]  69  16      1         10        1   140
#> [383,]  69   8      1          1        0   164
#> [384,]  72  16      1          1        1   130
#> [385,]  70   4      0        180        0   180
#> [386,]  69   1      1          0        0   155
#> [387,]  73   6      1          0        1   270
#> [388,]  73   7      0          7        1   140
#> [389,]  70   3      0          3        1   159
#> [390,]  70  13      1          9        0   100
#> [391,]  72   6      0        180        1   130
#> [392,]  73   4      0        180        1   154
#> [393,]  69   2      1          0        1   110
#> [394,]  71   3      1          1        0   150
#> [395,]  74  20      0         20        1   180
#> [396,]  68   9      0        180        1   120
#> [397,]  70   5      1          0        1   190
#> [398,]  71   3      1          2        1   190
#> [399,]  73  10      1          8        0   106
#> [400,]  69  12      1          1        1   149
#> [401,]  74   4      0          4        0   120
#> [402,]  73   4      0         58        1   160
#> [403,]  72   5      1          3        1   160
#> [404,]  73   6      0        180        0   110
#> [405,]  72  15      1          0        1   150
#> [406,]  71   7      1          2        0   143
#> [407,]  72   8      1          0        1   140
#> [408,]  74   3      0          3        1   150
#> [409,]  73  17      1         11        0   140
#> [410,]  71  13      1          8        0   121
#> [411,]  70   4      1          0        1   140
#> [412,]  72  15      1         13        0   156
#> [413,]  70   8      0          8        0   120
#> [414,]  73  10      1          9        1   146
#> [415,]  73  10      1         10        1   120
#> [416,]  74  15      1          9        1   179
#> [417,]  71   2      0         10        1   112
#> [418,]  73   1      0          1        1    80
#> [419,]  75   9      1          7        0   140
#> [420,]  75  13      1          1        1   130
#> [421,]  71   4      0          4        0   134
#> [422,]  72  15      1         12        1   120
#> [423,]  70   7      1          4        0   184
#> [424,]  72   1      1          1        0   168
#> [425,]  72   7      0         57        1   145
#> [426,]  73  10      0        180        0   162
#> [427,]  72  11      0         11        1   140
#> [428,]  70   3      0          3        0   150
#> [429,]  73  12      1         12        1   140
#> [430,]  72   2      0        180        0   120
#> [431,]  72   4      1          0        1   197
#> [432,]  73   4      0        180        0   124
#> [433,]  76   3      1          0        1   120
#> [434,]  71  32      1         12        1   107
#> [435,]  72   3      0        180        0   160
#> [436,]  77   4      0          4        0   185
#> [437,]  75   3      1          1        0   180
#> [438,]  72   7      1          2        0   142
#> [439,]  73  15      0         15        1   160
#> [440,]  71  16      0        180        0   140
#> [441,]  73  10      1         10        0   124
#> [442,]  74   7      0        180        1   150
#> [443,]  74   3      0          3        1   128
#> [444,]  74   2      1          1        0   140
#> [445,]  74  19      1          4        1   200
#> [446,]  73   6      0          6        1   114
#> [447,]  74   2      0        180        0   190
#> [448,]  72   4      1          3        0   160
#> [449,]  76  17      1          0        1   200
#> [450,]  76  13      1         10        0   110
#> [451,]  75   4      1          0        1   122
#> [452,]  75   7      0          7        0   190
#> [453,]  73  13      1         11        0   195
#> [454,]  75  12      0         12        1   160
#> [455,]  74   8      1          0        1   105
#> [456,]  76  13      1          8        1   148
#> [457,]  74   6      0        180        0   160
#> [458,]  75   1      0          1        1   125
#> [459,]  73   1      0         52        1   105
#> [460,]  73   0      0        180        0   156
#> [461,]  72   5      0        180        0   120
#> [462,]  78  12      1         11        1   160
#> [463,]  76  44      1         10        0   105
#> [464,]  76   5      0        180        0   185
#> [465,]  74  10      1          0        1   135
#> [466,]  73  33      1         12        1   175
#> [467,]  77   5      1          0        0   123
#> [468,]  77  12      1          9        1   100
#> [469,]  73  10      1          9        0   146
#> [470,]  77  12      0        180        0   130
#> [471,]  78   5      1          0        1   170
#> [472,]  75   3      1          1        1   171
#> [473,]  74   9      1          8        0   200
#> [474,]  79  10      1          8        0   190
#> [475,]  74   2      1          0        1   130
#> [476,]  78  18      0         18        1   144
#> [477,]  77   3      0        180        0   110
#> [478,]  76  29      0         47        0    90
#> [479,]  73   8      1          1        1   162
#> [480,]  73  11      1          2        1   110
#> [481,]  74   2      0        180        0   100
#> [482,]  78   8      1          6        1   110
#> [483,]  74   7      0          7        0   161
#> [484,]  76  13      1          1        1   170
#> [485,]  78  32      1          9        1   198
#> [486,]  80  10      1          6        1   147
#> [487,]  78   0      0        180        1   212
#> [488,]  75   5      0        119        1   150
#> [489,]  75  12      1          1        1   120
#> [490,]  78  15      0        180        1   270
#> [491,]  75  13      1          6        0   150
#> [492,]  74  10      1          8        0   135
#> [493,]  76   1      0          1        1    83
#> [494,]  79   4      0         80        0   145
#> [495,]  78   2      1          1        0   130
#> [496,]  75   4      1          0        0   212
#> [497,]  77   2      1          0        1   143
#> [498,]  76  11      1          0        0   120
#> [499,]  75   3      0          3        0     0
#> [500,]  76   7      0         29        1   150
#> [501,]  77  24      0         24        1   160
#> [502,]  78   6      1          0        1   240
#> [503,]  76   3      1          0        1   140
#> [504,]  78  11      1          1        1   140
#> [505,]  79  11      0        180        0   160
#> [506,]  79   2      1          0        1   121
#> [507,]  78  14      1          0        1   140
#> [508,]  81   1      0          1        0   130
#> [509,]  78  11      1          8        1   118
#> [510,]  76   4      0          4        1   160
#> [511,]  79   4      0          4        1   125
#> [512,]  76  12      1         10        1   127
#> [513,]  77   6      0          6        1   107
#> [514,]  80   3      1          0        1   120
#> [515,]  78  11      0        180        1   135
#> [516,]  76   1      0          1        1   140
#> [517,]  77  31      1          3        1   161
#> [518,]  78   7      1          0        1   110
#> [519,]  77   7      0        180        1   170
#> [520,]  80  15      1         12        1   150
#> [521,]  80  40      1          0        1   138
#> [522,]  78   4      0         59        1   112
#> [523,]  80  17      1         12        0   100
#> [524,]  81   4      1          2        1   126
#> [525,]  80   9      0        118        1   186
#> [526,]  78  32      0        180        1   130
#> [527,]  79   1      0         37        1   140
#> [528,]  81   3      0        180        0   184
#> [529,]  81   2      0        175        0   172
#> [530,]  78   7      0          7        1   147
#> [531,]  78  15      0         15        0   165
#> [532,]  80   5      1          1        1   108
#> [533,]  78  26      1          5        0   194
#> [534,]  76   1      0        166        0   131
#> [535,]  78  20      1          0        1   109
#> [536,]  80   1      0          1        0   100
#> [537,]  78   3      1          1        1   152
#> [538,]  82   3      1          1        1   144
#> [539,]  77   5      0         85        0   188
#> [540,]  80   2      1          1        0   168
#> [541,]  80   5      0          5        1   130
#> [542,]  82   1      1          0        1    82
#> [543,]  79  10      0        180        1   150
#> [544,]  81   1      0        108        0   129
#> [545,]  82  21      1          2        0   155
#> [546,]  79   4      0          4        1   121
#> [547,]  80   6      0          6        1   110
#> [548,]  83   9      1          5        1   170
#> [549,]  82   5      0        180        0   110
#> [550,]  79   7      1          6        0   130
#> [551,]  81  11      1          8        0   160
#> [552,]  81   5      0        177        0    41
#> [553,]  80  11      1          8        0   170
#> [554,]  79   4      0          4        1   183
#> [555,]  78   9      1          4        1   120
#> [556,]  82   8      1          1        0   128
#> [557,]  79   1      0        180        1   170
#> [558,]  81  15      0        180        1   140
#> [559,]  80   7      1          0        1   146
#> [560,]  81  20      1          9        0   170
#> [561,]  81  16      0         16        1   110
#> [562,]  80  11      1          8        0   110
#> [563,]  80   8      1          7        0   160
#> [564,]  79   7      0        177        0   197
#> [565,]  79   0      1          0        1    96
#> [566,]  85   4      0        180        0    90
#> [567,]  81   2      1          1        0   198
#> [568,]  83   2      0          2        1   155
#> [569,]  84   4      0        167        0   198
#> [570,]  84   4      0          4        1    85
#> [571,]  84   1      0         38        1   205
#> [572,]  83   3      0        180        0   174
#> [573,]  79   9      1          8        0   150
#> [574,]  79   4      0          4        1    60
#> [575,]  80   6      0         71        1   189
#> [576,]  83   1      0          1        1   100
#> [577,]  80  30      1         13        0   220
#> [578,]  83   9      0        180        0   198
#> [579,]  79  14      1          0        0   110
#> [580,]  81  14      1         12        1   128
#> [581,]  83   2      0        154        0   130
#> [582,]  82   0      0          2        1   100
#> [583,]  85   9      1          6        1   160
#> [584,]  83   1      0        180        0   160
#> [585,]  81   4      0          4        0   175
#> [586,]  84  15      1         13        1   110
#> [587,]  81   1      0          1        1   145
#> [588,]  81  12      0         12        1   163
#> [589,]  82  16      1          8        0   103
#> [590,]  83  12      1          2        1   170
#> [591,]  81  19      1         14        0   120
#> [592,]  80   2      0         88        0   135
#> [593,]  83   7      0        126        0   135
#> [594,]  81  16      1          9        0   180
#> [595,]  84   6      0        165        0   145
#> [596,]  86   3      0          3        1   140
#> [597,]  82   9      0        180        1   134
#> [598,]  81  13      0        180        0   152
#> [599,]  81   2      1          0        1   118
#> [600,]  81   4      0        180        0   160
#> [601,]  83   9      0        180        1   149
#> [602,]  82   1      0        180        1   193
#> [603,]  86  12      1          0        1   132
#> [604,]  82  14      1         11        1   103
#> [605,]  86   6      1          0        1   140
#> [606,]  84   3      1          2        0   125
#> [607,]  83  10      1          0        1   190
#> [608,]  86   2      0        180        1   169
#> [609,]  84   3      0          3        1   121
#> [610,]  83  13      1         12        0   170
#> [611,]  84   7      1          2        0   148
#> [612,]  87   2      0        180        0   113
#> [613,]  84   9      0         92        1   110
#> [614,]  84   3      0        180        1   170
#> [615,]  82   4      0          4        0   130
#> [616,]  86   6      0          6        1   117
#> [617,]  84  13      0         62        1   100
#> [618,]  86   6      1          1        0   112
#> [619,]  88   4      0          4        0   100
#> [620,]  83  20      1          3        1   150
#> [621,]  88   4      0          4        1   115
#> [622,]  86   9      1          7        1   142
#> [623,]  87   2      0        180        1   130
#> [624,]  86   6      0         46        0   173
#> [625,]  88   3      0        115        0   110
#> [626,]  88   4      0          4        0    86
#> [627,]  87   1      0          1        0   170
#> [628,]  84   8      0        180        1   119
#> [629,]  85   8      0          8        1   136
#> [630,]  84   2      0        110        1   174
#> [631,]  87  15      1          9        1   138
#> [632,]  89  10      0         46        1   170
#> [633,]  90  14      0         14        1   100
#> [634,]  86   4      0        180        1   145
#> [635,]  91   8      0          8        0   100
#> [636,]  87   2      0        180        0   160
#> [637,]  87   6      1          0        0   125
#> [638,]  86   3      1          0        1    80
#> [639,]  88   7      0         24        0   119
#> [640,]  88   8      0         50        1   154
#> [641,]  87   6      0        126        1   168
#> [642,]  86  10      0        180        1   137
#> [643,]  86   9      1          7        0   130
#> [644,]  90   4      1          0        0   121
#> [645,]  91   1      0          1        1    74
#> [646,]  87  43      0        178        1   130
#> [647,]  87   5      0         36        1   150
#> [648,]  90   5      1          0        1   125
#> [649,]  88   3      1          2        0   159
#> [650,]  89   3      1          1        1   160
#> [651,]  89  12      1          0        1   130
#> [652,]  89   2      0        168        0   118
#> [653,]  92   7      0          7        1   110
#> [654,]  89   4      0          4        1   159
#> [655,]  89  14      0        180        1    84
#> [656,]  90  19      1         11        1   129
#> [657,]  94   6      0         50        0    78
#> [658,]  90   1      0          1        1   118
#> [659,]  93   8      0        179        1   110
#> [660,]  94   8      0          8        1   142
#> [661,]  92   4      0         76        1   149
#> [662,]  91   1      0        180        0   158
#> [663,]  96   3      0         12        1    97
#> [664,]  95   8      1          5        1   150
#> [665,]  91  12      0         53        1   212
#> [666,]  91   7      0          7        0   135
#> [667,]  93   0      1          0        1   122
#> [668,]  92   2      0          2        0   112
#> [669,]  93   4      0        180        1   135
#> [670,]  96  15      1          0        1   140
#> 
#> $y
#>   [1]   5.0+   5.0+ 180.0+   5.0+ 180.0+   2.0+ 180.0+ 180.0+  12.0    5.0+
#>  [11] 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 180.0+   5.0+ 180.0+ 180.0+   3.0 
#>  [21] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 180.0+ 155.0+ 180.0+ 180.0+
#>  [31] 180.0+ 180.0+ 180.0+ 180.0+ 150.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [41] 180.0+ 180.0+ 180.0+  73.0  161.0+ 180.0+ 180.0+ 180.0+   5.0+ 180.0+
#>  [51] 180.0+ 177.0+ 180.0+ 180.0+ 180.0+ 180.0+  10.0+ 172.0+ 180.0+ 180.0+
#>  [61] 180.0+   7.0  180.0+   2.0  179.0+ 179.0+ 180.0+ 180.0+   4.0+ 180.0+
#>  [71] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   7.0    4.0+
#>  [81]  77.0  180.0+ 180.0+ 180.0+   5.0  180.0+ 180.0+ 180.0+ 180.0+   4.0+
#>  [91]  85.0  166.0+  16.0+ 180.0+ 152.0+   7.0+   6.0+ 180.0+ 180.0+ 180.0+
#> [101]  13.0+ 180.0+ 174.0+   6.0+   1.0    9.0+ 180.0+ 180.0+ 175.0+ 180.0+
#> [111]   2.0  180.0+ 180.0+ 180.0+ 180.0+  16.0+ 180.0+  16.0  180.0+ 180.0+
#> [121] 180.0+  12.0+ 134.0+ 180.0+   8.0    2.0  180.0+ 180.0+ 180.0+ 180.0+
#> [131]   1.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+   8.0+ 180.0+ 180.0+   0.5 
#> [141] 180.0+ 180.0+ 180.0+ 180.0+ 171.0+  15.0  166.0+   5.0+ 180.0+ 147.0+
#> [151] 180.0+ 180.0+ 180.0+   4.0+ 180.0+   9.0+ 180.0+   2.0+ 180.0+ 180.0+
#> [161] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [171]   9.0+ 180.0+   0.5  180.0+ 171.0+ 180.0+ 180.0+   3.0    1.0  180.0+
#> [181] 180.0+  10.0+ 180.0+ 180.0+  17.0   45.0    3.0+  19.0  180.0+   9.0+
#> [191] 180.0+ 172.0+ 172.0+   1.0+ 180.0+ 180.0+ 180.0+  13.0+ 180.0+ 180.0+
#> [201] 180.0+ 170.0   94.0    6.0  180.0+ 180.0+   3.0+ 180.0+ 180.0+  28.0 
#> [211]  13.0+   5.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   9.0 
#> [221]   7.0+  22.0  180.0+ 180.0+  84.0    7.0+ 180.0+   1.0    1.0    4.0+
#> [231] 167.0    6.0+ 180.0+ 180.0+  17.0  180.0+  12.0  180.0+ 180.0+ 180.0+
#> [241]  14.0+  36.0  180.0+   3.0+  12.0  180.0+ 180.0+ 180.0+ 180.0+  12.0+
#> [251] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+
#> [261] 180.0+ 180.0+ 180.0+ 180.0+   3.0+   2.0+ 103.0    5.0+  13.0  180.0+
#> [271] 179.0+ 166.0+  14.0+   3.0  180.0+ 180.0+ 175.0+ 180.0+ 180.0+   7.0+
#> [281]   8.0    5.0  180.0+   1.0  180.0+ 180.0+   1.0+  18.0   11.0+ 180.0+
#> [291]  79.0   15.0+   4.0+  15.0  180.0+ 180.0+ 180.0+ 180.0+   8.0+   3.0 
#> [301]  10.0  180.0+ 180.0+ 180.0+  19.0+  12.0  180.0+ 179.0+ 180.0+  11.0+
#> [311]   0.5  180.0+ 152.0+ 180.0+  93.0   21.0+  18.0+   4.0    7.0+ 150.0 
#> [321] 180.0+ 180.0+ 180.0+ 174.0+   6.0  180.0+   0.5  180.0+ 180.0+  14.0+
#> [331]  45.0  103.0  180.0+  36.0    5.0+ 180.0+ 180.0+  97.0  180.0+   7.0 
#> [341]   8.0+   2.0+ 180.0+ 172.0+ 180.0+   7.0  180.0+   8.0+  13.0+ 123.0 
#> [351] 180.0+ 180.0+  51.0  180.0+  60.0   76.0  180.0+ 132.0   10.0+ 180.0+
#> [361] 162.0    7.0+   7.0+   9.0  180.0+ 180.0+ 180.0+  12.0  180.0+ 180.0+
#> [371] 180.0+ 180.0+   2.0   76.0  173.0+ 180.0+ 180.0+ 180.0+ 180.0+  28.0 
#> [381] 180.0+  16.0+ 180.0+  16.0+ 180.0+ 180.0+   6.0    7.0+   3.0+  13.0+
#> [391] 180.0+ 180.0+   2.0    3.0+  20.0  180.0+ 180.0+   3.0   87.0   12.0 
#> [401]   4.0+  58.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+   3.0  180.0+ 175.0 
#> [411] 180.0+ 180.0+   8.0+ 180.0+  15.0  180.0+  10.0    1.0  180.0+  13.0 
#> [421]   4.0+ 180.0+ 104.0+   1.0   57.0  180.0+  11.0    3.0+  12.0  180.0+
#> [431] 180.0+ 180.0+ 180.0+ 177.0+ 180.0+   4.0+ 180.0+   7.0   15.0+ 180.0+
#> [441]  10.0  180.0+   3.0  180.0+ 180.0+   6.0  180.0+ 180.0+  17.0+ 174.0+
#> [451]   4.0    7.0  180.0+  12.0  180.0+ 180.0+ 180.0+   1.0   52.0  180.0+
#> [461] 180.0+  12.0  180.0+ 180.0+ 180.0+  33.0    5.0  180.0+ 180.0+ 180.0+
#> [471] 180.0+   3.0  168.0+ 180.0+ 176.0+  18.0  180.0+  47.0  180.0+  11.0 
#> [481] 180.0+   8.0+   7.0  180.0+  32.0   10.0  180.0+ 119.0   12.0  180.0+
#> [491] 180.0+ 180.0+   1.0   80.0  180.0+   4.0+   2.0   11.0    3.0   29.0 
#> [501]  24.0  180.0+   3.0+ 180.0+ 180.0+ 180.0+ 180.0+   1.0   11.0    4.0 
#> [511]   4.0  180.0+   6.0    3.0+ 180.0+   1.0  171.0   43.0  180.0+ 180.0+
#> [521]  40.0   59.0   17.0   93.0  118.0  180.0+  37.0  180.0+ 175.0+   7.0+
#> [531]  15.0+   5.0+ 171.0+ 166.0+  20.0+   1.0    3.0+ 180.0+  85.0   10.0 
#> [541]   5.0    1.0  180.0+ 108.0  180.0+   4.0    6.0    9.0+ 180.0+ 180.0+
#> [551] 180.0+ 177.0+ 169.0    4.0  180.0+ 180.0+ 180.0+ 180.0+   7.0+  20.0 
#> [561]  16.0  180.0+ 180.0+ 177.0+   0.5  180.0+ 180.0+   2.0  167.0    4.0 
#> [571]  38.0  180.0+ 180.0+   4.0   71.0    1.0   30.0  180.0+ 180.0+ 180.0+
#> [581] 154.0    2.0  180.0+ 180.0+   4.0+ 180.0+   1.0   12.0   16.0+  77.0 
#> [591] 180.0+  88.0  126.0  180.0+ 165.0    3.0  180.0+ 180.0+ 180.0+ 180.0+
#> [601] 180.0+ 180.0+ 180.0+ 174.0    6.0  180.0+ 180.0+ 180.0+   3.0   13.0 
#> [611] 180.0+ 180.0+  92.0  180.0+   4.0    6.0+  62.0    6.0+   4.0+  20.0 
#> [621]   4.0   11.0  180.0+  46.0  115.0    4.0    1.0+ 180.0+   8.0  110.0 
#> [631] 180.0+  46.0   14.0  180.0+   8.0  180.0+  25.0    3.0   24.0   50.0 
#> [641] 126.0  180.0+ 180.0+   4.0    1.0  178.0+  36.0   89.0   75.0    3.0+
#> [651] 180.0+ 168.0    7.0    4.0  180.0+ 180.0+  50.0    1.0+ 179.0+   8.0+
#> [661]  76.0  180.0+  12.0    8.0   53.0    7.0+   0.5    2.0  180.0+  15.0+
#> 
#> $weights
#> NULL
#> 
#> $offset
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8388156