Cross-Validated GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::cv.glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.cv.glmnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.cv.glmnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.cv_glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

foldid	untyped	NULL		-

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nfolds	integer	10		\([3, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	TRUE	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	lambda.1se		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

standardize.response	logical	FALSE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.gaussian	character	-	covariance, naive	-

type.logistic	character	Newton	Newton, modified.Newton	-

type.measure	character	deviance	deviance, C	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.coxnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvCVGlmnet

Methods

Public methods

LearnerSurvCVGlmnet$new()

LearnerSurvCVGlmnet$selected_features()

LearnerSurvCVGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvCVGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvCVGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvCVGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("surv.cv_glmnet")
print(learner)
#> 
#> ── <LearnerSurvCVGlmnet> (surv.cv_glmnet): Regularized Generalized Linear Model 
#> • Model: -
#> • Parameters: use_pred_offset=TRUE
#> • Packages: mlr3, mlr3proba, mlr3extralearners, and glmnet
#> • Predict Types: [crank], distr, and lp
#> • Feature Types: logical, integer, and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: offset, selected_features, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("grace")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#> Measure: Partial Likelihood Deviance 
#> 
#>      Lambda Index Measure     SE Nonzero
#> min 0.00345    44   2.734 0.1509       6
#> 1se 0.07426    11   2.866 0.1525       3
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  28   9      0        180        1   107
#>   [2,]  32   5      1          0        1   121
#>   [3,]  33   2      0          2        0   150
#>   [4,]  35   5      1          2        0   172
#>   [5,]  35  10      1          9        0   106
#>   [6,]  34   5      0          5        0   120
#>   [7,]  35   2      1          1        1   112
#>   [8,]  37   9      0        180        1   151
#>   [9,]  38  13      1          0        1   161
#>  [10,]  38   2      0        115        0   150
#>  [11,]  36   1      0        180        1   155
#>  [12,]  35   0      0        180        1   119
#>  [13,]  36   5      1          0        1   115
#>  [14,]  33   6      1          1        1   115
#>  [15,]  38  16      1         10        0   160
#>  [16,]  38  12      1         11        1    92
#>  [17,]  42   3      1          1        1   130
#>  [18,]  37   1      1          0        1   146
#>  [19,]  42   2      0        180        1   100
#>  [20,]  38   5      1          3        0   125
#>  [21,]  43   3      1          0        1   100
#>  [22,]  40   1      1          0        1   145
#>  [23,]  42   4      0        180        0   162
#>  [24,]  42  12      1         10        1   170
#>  [25,]  43   2      1          1        1   116
#>  [26,]  42   2      0        180        1   124
#>  [27,]  41  10      1          8        0   150
#>  [28,]  41  13      1          1        0   140
#>  [29,]  45   6      0        180        1   170
#>  [30,]  41   5      1          4        1   141
#>  [31,]  45   2      0        180        1   140
#>  [32,]  46   2      1          1        0   126
#>  [33,]  47   4      1          3        0   118
#>  [34,]  48  15      0        180        1   160
#>  [35,]  45   3      0        150        0   130
#>  [36,]  44   3      1          0        1   180
#>  [37,]  43  29      0        180        1   180
#>  [38,]  45   4      1          0        0   124
#>  [39,]  43  10      0        180        0   185
#>  [40,]  47   6      1          0        1   116
#>  [41,]  46  13      1         10        0   100
#>  [42,]  43   3      1          0        1   124
#>  [43,]  45   8      1          0        1   117
#>  [44,]  49   5      0         73        1   136
#>  [45,]  45   5      0          5        0   141
#>  [46,]  46   2      1          1        1   122
#>  [47,]  44   4      1          0        1   114
#>  [48,]  47   2      0        180        0   108
#>  [49,]  46   5      1          3        0   130
#>  [50,]  46   4      0        180        1   121
#>  [51,]  44   2      0        180        0   142
#>  [52,]  46  15      0        180        1   120
#>  [53,]  45   9      1          0        1   145
#>  [54,]  47   3      1          1        1   120
#>  [55,]  48   3      0        180        0   154
#>  [56,]  48  12      1         11        0   200
#>  [57,]  47   5      1          3        1   130
#>  [58,]  47   9      1          6        0   170
#>  [59,]  49   4      0        180        0   117
#>  [60,]  50   1      1          0        1   129
#>  [61,]  48   2      1          0        0   184
#>  [62,]  50   4      1          1        0   125
#>  [63,]  50   6      1          2        1   140
#>  [64,]  46   3      1          1        1   140
#>  [65,]  50   7      0        180        1   110
#>  [66,]  49   2      0          2        0   105
#>  [67,]  49  15      1         11        1   160
#>  [68,]  47   2      0        180        0   150
#>  [69,]  49  23      0        179        1   112
#>  [70,]  52   2      0        180        1   170
#>  [71,]  50   7      1          0        1    92
#>  [72,]  50   4      0          4        1   100
#>  [73,]  51   3      1          2        0   113
#>  [74,]  50   1      1          0        0   150
#>  [75,]  50   9      0        180        0   130
#>  [76,]  49   7      1          4        1    90
#>  [77,]  47   8      0        180        0   160
#>  [78,]  47   6      0        180        1   162
#>  [79,]  52   2      0        180        0   155
#>  [80,]  46   3      0        180        1   120
#>  [81,]  46   1      1          1        0   145
#>  [82,]  50   4      1          1        0   150
#>  [83,]  48   7      1          0        1   110
#>  [84,]  47   2      1          1        0   110
#>  [85,]  52   4      1          4        0   152
#>  [86,]  49   9      1          3        0   102
#>  [87,]  49  15      0        180        1   160
#>  [88,]  53   5      0        180        1   140
#>  [89,]  53   5      0         77        0   159
#>  [90,]  53   7      1          0        0   199
#>  [91,]  54   6      1          3        0   129
#>  [92,]  50   2      0          5        1   106
#>  [93,]  50  14      1         13        0   170
#>  [94,]  53   8      1          7        0   160
#>  [95,]  51  25      1          1        0   202
#>  [96,]  49   5      1          2        1   150
#>  [97,]  53   4      0          4        0   140
#>  [98,]  52  14      1          7        1   200
#>  [99,]  48   6      0        180        0   160
#> [100,]  48  11      1         10        0   120
#> [101,]  51  13      0         99        1   160
#> [102,]  49  16      0         16        0   125
#> [103,]  55   3      1          1        0   150
#> [104,]  54  23      1         10        0   131
#> [105,]  52   7      1          2        0   154
#> [106,]  55   6      1          2        1   114
#> [107,]  54   9      1          1        0   130
#> [108,]  55   4      1          2        0   150
#> [109,]  52   4      0        180        1   180
#> [110,]  51  13      1         11        0   145
#> [111,]  50   3      0        174        1   153
#> [112,]  55  28      1         13        1   160
#> [113,]  49   6      1          0        1   130
#> [114,]  49   1      0          1        1   110
#> [115,]  50   7      1          0        1   127
#> [116,]  56   4      1          1        1   130
#> [117,]  55   1      0        180        0   127
#> [118,]  54   1      0        180        0   162
#> [119,]  54   7      1          0        1   100
#> [120,]  56   2      0        180        0   132
#> [121,]  55   5      1          4        1   120
#> [122,]  53  18      1          9        1   150
#> [123,]  55   6      0        180        0   170
#> [124,]  53  10      1          9        0   172
#> [125,]  52  16      1         14        0   170
#> [126,]  53  15      0         15        1    90
#> [127,]  53   4      0        180        1   150
#> [128,]  55   6      0        180        1   100
#> [129,]  55   6      1          5        1   138
#> [130,]  54  12      1          0        1   190
#> [131,]  56   3      0          8        1   139
#> [132,]  55   1      0          2        0   130
#> [133,]  54   7      1          2        0   129
#> [134,]  54   2      1          1        1   176
#> [135,]  57   5      1          3        1   138
#> [136,]  57   1      0          1        1   100
#> [137,]  56   4      1          0        1   140
#> [138,]  52  15      1         14        0   130
#> [139,]  56  14      1         11        0   130
#> [140,]  53   3      1          0        1   200
#> [141,]  57  10      0        180        1   170
#> [142,]  58   8      0          8        1   130
#> [143,]  54   5      0        180        1   108
#> [144,]  55   3      1          1        1   156
#> [145,]  57   0      0          0        1   150
#> [146,]  53  21      1         13        1   130
#> [147,]  59   3      1          1        0   172
#> [148,]  57   4      0        180        1   119
#> [149,]  58   6      1          0        1    90
#> [150,]  53  15      1         10        1   130
#> [151,]  54  17      1          8        1   227
#> [152,]  55   9      1          2        1   147
#> [153,]  55  13      0        166        1   140
#> [154,]  57   4      1          2        1   185
#> [155,]  53   4      0        147        1   145
#> [156,]  53   7      1          0        1   120
#> [157,]  57  11      1         10        1   129
#> [158,]  55   3      1          2        0   140
#> [159,]  55   5      0          5        1   131
#> [160,]  54   7      1          0        1   141
#> [161,]  56   4      0          4        0   164
#> [162,]  59  15      1         10        0   140
#> [163,]  58   9      1          0        1   180
#> [164,]  58   1      1          1        1   200
#> [165,]  55   5      1          0        0   140
#> [166,]  55   2      0          2        0   106
#> [167,]  59   9      1          1        1   125
#> [168,]  57   1      0        180        0   148
#> [169,]  60  11      1          9        0   106
#> [170,]  59   3      0        180        0   120
#> [171,]  58   4      1          0        1   160
#> [172,]  57   2      0          2        1   120
#> [173,]  60   5      1          1        0   138
#> [174,]  57   5      0        180        1   130
#> [175,]  58  11      1          9        1   124
#> [176,]  55   5      1          0        1   160
#> [177,]  59   5      0        180        1   155
#> [178,]  59   4      1          0        1   152
#> [179,]  58  26      1          0        1   189
#> [180,]  58   4      1          3        0   120
#> [181,]  58   8      0        161        1   140
#> [182,]  61   4      1          3        0   151
#> [183,]  61   9      1          8        0   150
#> [184,]  61   3      1          2        1   102
#> [185,]  58   1      0          1        1   100
#> [186,]  61  20      1         13        0   130
#> [187,]  57   2      1          0        1   116
#> [188,]  58  10      0         10        1   150
#> [189,]  57  11      0        180        1   150
#> [190,]  56  14      0         45        0   130
#> [191,]  57   3      1          2        0   120
#> [192,]  56  13      1          6        1   158
#> [193,]  55   4      1          3        1   160
#> [194,]  58  11      0        172        1   135
#> [195,]  60  12      1          0        1   114
#> [196,]  61   4      1          0        1   115
#> [197,]  56   8      1          8        0   120
#> [198,]  59  11      1          8        1   190
#> [199,]  57  15      1         13        1   110
#> [200,]  59   5      1          2        0   182
#> [201,]  58   5      1          1        1   135
#> [202,]  59  10      0        180        0   160
#> [203,]  61   8      0         77        0   120
#> [204,]  61  13      0         13        0   210
#> [205,]  62   7      1          2        1   180
#> [206,]  57   3      1          0        0   100
#> [207,]  61  18      0        170        0   140
#> [208,]  57   7      0        169        0   180
#> [209,]  61   7      0          7        1   150
#> [210,]  60   7      0          7        0   147
#> [211,]  61   6      0          6        0   134
#> [212,]  59  13      1          2        0   198
#> [213,]  57  12      1          9        1   120
#> [214,]  62   4      1          0        0   160
#> [215,]  60  17      1          8        1   140
#> [216,]  58   3      1          0        1   146
#> [217,]  62   4      1          3        0   173
#> [218,]  59  16      1          9        1   133
#> [219,]  63   6      0         28        1   120
#> [220,]  58  11      1          9        0   179
#> [221,]  57   2      1          1        0   159
#> [222,]  62  17      1         10        1   180
#> [223,]  58   7      0        180        1   150
#> [224,]  63   4      1          3        0   222
#> [225,]  63   4      0        180        1   190
#> [226,]  63   4      1          1        0   155
#> [227,]  60  18      1         13        0   132
#> [228,]  59   8      0        180        1   140
#> [229,]  61   9      1          9        1   150
#> [230,]  58   9      1          9        0   110
#> [231,]  62   7      0          7        0   150
#> [232,]  59   1      0         22        1   162
#> [233,]  58   2      0        180        0   127
#> [234,]  60   7      1          5        1   141
#> [235,]  59   5      1          1        0   148
#> [236,]  60   7      1          1        1    90
#> [237,]  65  13      0        180        1   100
#> [238,]  63   1      0          1        0   162
#> [239,]  63   1      0          1        0   130
#> [240,]  62   6      0        180        0   170
#> [241,]  61  15      1         13        0   170
#> [242,]  59   4      0          4        0   149
#> [243,]  60   3      0          3        0   168
#> [244,]  64  10      1          9        0   160
#> [245,]  62   6      0          6        0   120
#> [246,]  63  12      1         10        0   200
#> [247,]  59  10      0        180        1   130
#> [248,]  61   6      1          1        1   117
#> [249,]  64  12      1         11        0   160
#> [250,]  63  10      1          0        1   148
#> [251,]  65  36      1         11        0   140
#> [252,]  66   3      1          1        0   127
#> [253,]  61  10      1          2        1   194
#> [254,]  63  12      1          9        0   114
#> [255,]  63   7      0        180        0   120
#> [256,]  65  10      1          8        1   120
#> [257,]  64   0      0          0        1    90
#> [258,]  64  21      1         10        0   190
#> [259,]  61  12      1         11        0   154
#> [260,]  65   3      0        180        1   190
#> [261,]  63  16      1          7        1   110
#> [262,]  66   6      1          1        1   130
#> [263,]  63  12      0         12        1   150
#> [264,]  62   3      1          1        1   199
#> [265,]  65   6      0          9        0   112
#> [266,]  65   3      1          0        1    80
#> [267,]  63   5      1          4        0   170
#> [268,]  63   2      1          1        0   180
#> [269,]  62  13      1         11        0   180
#> [270,]  64   2      0          2        0   201
#> [271,]  66  16      1         11        1   169
#> [272,]  62   9      0        180        0   145
#> [273,]  63   9      1          8        1   160
#> [274,]  63   3      1          2        0   120
#> [275,]  63   2      1          0        0   140
#> [276,]  64  19      1          8        1   160
#> [277,]  65   8      1          0        1   140
#> [278,]  65  15      1         11        1   160
#> [279,]  64  13      1         12        1   150
#> [280,]  66   7      1          0        1   115
#> [281,]  64  14      1         13        1   150
#> [282,]  65   3      0          3        0   105
#> [283,]  64   0      0          0        1   148
#> [284,]  67   4      1          3        0   130
#> [285,]  66   6      1          0        1   140
#> [286,]  65   2      1          1        1   170
#> [287,]  68   1      0        180        1   166
#> [288,]  64  10      1          9        1   110
#> [289,]  67   8      1          1        1   130
#> [290,]  68   5      0          5        1    90
#> [291,]  63  10      0         16        1   160
#> [292,]  64   1      0          1        1   120
#> [293,]  68  18      0        180        1   260
#> [294,]  63   8      1          1        1   162
#> [295,]  65  18      1          3        0   120
#> [296,]  63   1      1          0        1   155
#> [297,]  68  11      0        180        0   160
#> [298,]  66  11      1          0        0   100
#> [299,]  69  12      0         15        1   140
#> [300,]  66  15      1         13        1   160
#> [301,]  63   2      0        180        0   150
#> [302,]  69  21      1         10        0   180
#> [303,]  69   6      0        180        1   100
#> [304,]  66   9      1          8        0   130
#> [305,]  63   8      0        180        1   120
#> [306,]  68  14      1         13        1   140
#> [307,]  65   8      1          0        1    90
#> [308,]  69   1      1          0        0   170
#> [309,]  67   1      0        180        1   160
#> [310,]  65   1      1          0        0   133
#> [311,]  63   2      1          0        0    99
#> [312,]  65   6      0          6        0    80
#> [313,]  66  19      1         12        1   150
#> [314,]  67  12      1         12        0   160
#> [315,]  69   6      0         99        1   140
#> [316,]  65   4      1          1        0   130
#> [317,]  64   4      0        179        0   160
#> [318,]  64  11      0         11        0   125
#> [319,]  64   0      1          0        1   118
#> [320,]  67   2      0         18        0   131
#> [321,]  66   4      0        180        0   177
#> [322,]  68   4      1          0        1   160
#> [323,]  69   4      1          3        1   150
#> [324,]  65   1      0          1        1   120
#> [325,]  68  18      1          0        1   160
#> [326,]  65   6      0        101        1   115
#> [327,]  68   4      0          4        1   190
#> [328,]  71   3      0          5        0   112
#> [329,]  70   7      1          0        1   190
#> [330,]  71  20      1          0        1   160
#> [331,]  67   2      0        180        0   128
#> [332,]  66   9      1          3        1   151
#> [333,]  66   1      1          1        1   165
#> [334,]  70   4      1          0        1   180
#> [335,]  67   6      1          4        0   130
#> [336,]  68  18      1         14        1   170
#> [337,]  69   0      0          0        1   148
#> [338,]  65   2      0        180        0   130
#> [339,]  69   3      1          2        0   151
#> [340,]  67  14      1         13        0   130
#> [341,]  65  14      1         13        1   150
#> [342,]  69   5      0          5        1   142
#> [343,]  71   3      0        103        0   133
#> [344,]  69   3      0          3        1   130
#> [345,]  70  22      1         13        0   103
#> [346,]  67   5      0          5        0   130
#> [347,]  69   8      1          5        1   195
#> [348,]  69   8      1          7        1   108
#> [349,]  67   3      0        180        0   110
#> [350,]  66   2      1          1        0   123
#> [351,]  69  19      0        180        0   130
#> [352,]  67  14      0        172        1   140
#> [353,]  67   7      1          4        0   122
#> [354,]  68   2      0          7        1   130
#> [355,]  69   8      1          2        0   121
#> [356,]  67  13      1          9        0   130
#> [357,]  70   9      0        180        1   142
#> [358,]  67  22      1          1        1   140
#> [359,]  68   3      0         19        0   135
#> [360,]  67  12      1          8        0   120
#> [361,]  67   1      0          1        1    60
#> [362,]  67   4      0         60        1   136
#> [363,]  69   5      0         76        0   120
#> [364,]  67   8      1          0        1   130
#> [365,]  72  13      1         11        1   195
#> [366,]  68  10      1          8        1   160
#> [367,]  66  24      1         13        0   130
#> [368,]  70  35      1          0        1   105
#> [369,]  72  30      1          0        1   145
#> [370,]  68   7      1          2        0   135
#> [371,]  71   6      0          9        0   120
#> [372,]  69  10      1          6        1   120
#> [373,]  70  11      0        180        1   210
#> [374,]  72  19      1          8        0   120
#> [375,]  72  12      1         10        0   170
#> [376,]  67   9      0        180        0   158
#> [377,]  72   2      0          2        1   100
#> [378,]  71   1      0        173        1   188
#> [379,]  68  23      0        180        1   220
#> [380,]  70   3      0        180        0   121
#> [381,]  71   3      1          2        0   150
#> [382,]  68   4      1          3        0   210
#> [383,]  72   5      0         28        0   120
#> [384,]  69  16      1         10        1   140
#> [385,]  69   8      1          1        0   164
#> [386,]  69   1      1          0        0   155
#> [387,]  73   6      1          0        1   270
#> [388,]  72   8      1          1        1   150
#> [389,]  71   2      1          0        1   180
#> [390,]  73   7      0          7        1   140
#> [391,]  70   3      0          3        1   159
#> [392,]  72   6      0        180        1   130
#> [393,]  74   8      1          0        1    85
#> [394,]  73   4      0        180        1   154
#> [395,]  69   2      1          0        1   110
#> [396,]  71   3      1          1        0   150
#> [397,]  71  15      1         11        0   165
#> [398,]  74  20      0         20        1   180
#> [399,]  68   9      0        180        1   120
#> [400,]  71  20      1         10        0   140
#> [401,]  74   0      1          0        1    90
#> [402,]  70   5      1          0        1   190
#> [403,]  71  17      1         11        0   160
#> [404,]  71   8      1          7        0   149
#> [405,]  74   4      0          4        0   120
#> [406,]  73   4      0         58        1   160
#> [407,]  72   5      1          3        1   160
#> [408,]  72  15      1          0        1   150
#> [409,]  71   7      1          2        0   143
#> [410,]  74   3      0          3        1   150
#> [411,]  73  17      1         11        0   140
#> [412,]  69   2      1          1        1    80
#> [413,]  70   4      1          0        1   140
#> [414,]  71  14      1         13        1   170
#> [415,]  74   7      1          0        1   117
#> [416,]  69   7      0        180        1   144
#> [417,]  72  15      1         13        0   156
#> [418,]  71  10      1          9        1   120
#> [419,]  75   1      0          1        0   133
#> [420,]  75   2      1          1        0   145
#> [421,]  73  10      1         10        1   120
#> [422,]  74  15      1          9        1   179
#> [423,]  71   2      0         10        1   112
#> [424,]  73   1      0          1        1    80
#> [425,]  75   9      1          7        0   140
#> [426,]  71   4      0          4        0   134
#> [427,]  72  15      1         12        1   120
#> [428,]  73  10      1          8        0   120
#> [429,]  70   7      1          4        0   184
#> [430,]  72   7      0         57        1   145
#> [431,]  73  10      0        180        0   162
#> [432,]  72  11      0         11        1   140
#> [433,]  70   3      0          3        0   150
#> [434,]  76  25      1         12        1   170
#> [435,]  73  12      1         12        1   140
#> [436,]  75   1      0        180        1   140
#> [437,]  72   4      1          0        1   197
#> [438,]  71   3      1          0        0   144
#> [439,]  73   5      0        180        0   126
#> [440,]  73   4      0        180        0   124
#> [441,]  74  34      1          8        1   233
#> [442,]  71  32      1         12        1   107
#> [443,]  76   5      0          5        1   130
#> [444,]  75   3      1          1        0   180
#> [445,]  72   7      1          2        0   142
#> [446,]  73  15      0         15        1   160
#> [447,]  73  10      1         10        0   124
#> [448,]  76   1      0        180        0   114
#> [449,]  74  19      1          4        1   200
#> [450,]  74   2      0        180        0   190
#> [451,]  72   4      1          3        0   160
#> [452,]  76  17      1          0        1   200
#> [453,]  76  13      1         10        0   110
#> [454,]  75   7      0          7        0   190
#> [455,]  73  13      1         11        0   195
#> [456,]  75  12      0         12        1   160
#> [457,]  74   8      1          0        1   105
#> [458,]  75   4      1          2        1   188
#> [459,]  74   6      0        180        0   160
#> [460,]  75   1      0          1        1   125
#> [461,]  74   2      0        180        0   111
#> [462,]  72   5      0        180        0   120
#> [463,]  78  12      1         11        1   160
#> [464,]  76  44      1         10        0   105
#> [465,]  76   5      0        180        0   185
#> [466,]  74  10      1          0        1   135
#> [467,]  76   5      1          0        1   167
#> [468,]  75   9      0        180        1   140
#> [469,]  73  33      1         12        1   175
#> [470,]  77   5      1          0        0   123
#> [471,]  77  12      1          9        1   100
#> [472,]  77   1      1          0        1    90
#> [473,]  73   7      1          0        0   174
#> [474,]  75   3      1          1        1   171
#> [475,]  75   6      0        180        0   150
#> [476,]  79  10      1          8        0   190
#> [477,]  74   2      1          0        1   130
#> [478,]  78  18      0         18        1   144
#> [479,]  77   3      0        180        0   110
#> [480,]  76  29      0         47        0    90
#> [481,]  73  11      1          2        1   110
#> [482,]  74   2      0        180        0   100
#> [483,]  78   7      0          7        1   133
#> [484,]  74  15      0        180        1   172
#> [485,]  78   8      1          6        1   110
#> [486,]  74   7      0          7        0   161
#> [487,]  78  32      1          9        1   198
#> [488,]  80  10      1          6        1   147
#> [489,]  78   0      0        180        1   212
#> [490,]  78  13      1          5        0   130
#> [491,]  75  12      1          1        1   120
#> [492,]  78  15      0        180        1   270
#> [493,]  80   8      0          8        1   120
#> [494,]  74  10      1          8        0   135
#> [495,]  78   2      1          1        0   130
#> [496,]  77   2      1          0        1   143
#> [497,]  78  10      0        180        1   130
#> [498,]  76  11      1          0        0   120
#> [499,]  76   7      0         29        1   150
#> [500,]  77  24      0         24        1   160
#> [501,]  80   9      0         23        1   128
#> [502,]  80   6      0          6        1   150
#> [503,]  78   6      1          0        1   240
#> [504,]  76   3      1          0        1   140
#> [505,]  79  11      0        180        0   160
#> [506,]  79   2      1          0        1   121
#> [507,]  78  14      1          0        1   140
#> [508,]  81   1      0          1        0   130
#> [509,]  78  11      1          8        1   118
#> [510,]  76   4      0          4        1   160
#> [511,]  79   4      0          4        1   125
#> [512,]  76  10      1          8        0   180
#> [513,]  76  12      1         10        1   127
#> [514,]  80   3      1          0        1   120
#> [515,]  75   2      1          1        1   204
#> [516,]  78  11      0        180        1   135
#> [517,]  76   1      0          1        1   140
#> [518,]  76   1      0          1        1    90
#> [519,]  79   3      0          3        0   120
#> [520,]  77   7      0        180        1   170
#> [521,]  77   6      0          6        1   144
#> [522,]  81   1      0        180        0   120
#> [523,]  80  15      1         12        1   150
#> [524,]  77   9      1          4        0   141
#> [525,]  82   5      0          8        1   120
#> [526,]  80  40      1          0        1   138
#> [527,]  76   7      0        161        0   151
#> [528,]  80  15      1          0        1    90
#> [529,]  79  28      0        164        0   100
#> [530,]  79   1      0         37        1   140
#> [531,]  81   3      0        180        0   184
#> [532,]  81   2      0        175        0   172
#> [533,]  77  13      1          0        1   190
#> [534,]  78  15      0         15        0   165
#> [535,]  80   5      1          1        1   108
#> [536,]  79   3      0          3        1   101
#> [537,]  78  26      1          5        0   194
#> [538,]  81   4      1          1        1   104
#> [539,]  78  20      1          0        1   109
#> [540,]  78   3      1          1        1   152
#> [541,]  77  10      1          8        1   130
#> [542,]  77   5      0         85        0   188
#> [543,]  80   2      1          1        0   168
#> [544,]  80   6      1          0        1   119
#> [545,]  80   5      0          5        1   130
#> [546,]  82   1      1          0        1    82
#> [547,]  79  10      0        180        1   150
#> [548,]  81   1      0        108        0   129
#> [549,]  78  12      0        180        0   134
#> [550,]  79   1      0        125        0   193
#> [551,]  82  21      1          2        0   155
#> [552,]  84  22      1         10        0   180
#> [553,]  79   4      0          4        1   121
#> [554,]  80   6      0          6        1   110
#> [555,]  83   9      1          5        1   170
#> [556,]  83   5      0        180        0   148
#> [557,]  79   7      1          6        0   130
#> [558,]  81  11      1          8        0   160
#> [559,]  80  11      1          8        0   170
#> [560,]  78  23      1         10        1   145
#> [561,]  79   4      0          4        1   183
#> [562,]  78   9      1          4        1   120
#> [563,]  81  15      0        180        1   140
#> [564,]  84   5      1          1        1    85
#> [565,]  81  20      1          9        0   170
#> [566,]  83   8      0          8        0   115
#> [567,]  81  16      0         16        1   110
#> [568,]  81   8      0        180        0   146
#> [569,]  79   7      0        177        0   197
#> [570,]  85   4      0        180        0    90
#> [571,]  81   2      1          1        0   198
#> [572,]  84   4      0        167        0   198
#> [573,]  80   3      1          1        1   230
#> [574,]  82  23      1          0        0   110
#> [575,]  84   5      0        180        1   203
#> [576,]  81   1      0          1        1   150
#> [577,]  84   1      0         38        1   205
#> [578,]  83   3      0        180        0   174
#> [579,]  81   4      0         90        1   138
#> [580,]  85   3      1          2        1   160
#> [581,]  84   4      0         89        1   129
#> [582,]  80   2      1          0        1   130
#> [583,]  80   6      0         71        1   189
#> [584,]  80  30      1         13        0   220
#> [585,]  79  14      1          0        0   110
#> [586,]  83   3      0        114        0    98
#> [587,]  81  14      1         12        1   128
#> [588,]  83   2      0        154        0   130
#> [589,]  85   9      1          6        1   160
#> [590,]  81   4      0          4        0   175
#> [591,]  81   1      0          1        1   145
#> [592,]  82  16      1          8        0   103
#> [593,]  86  12      0        180        1   120
#> [594,]  83  12      1          2        1   170
#> [595,]  81  19      1         14        0   120
#> [596,]  82   3      1          2        0   130
#> [597,]  80   2      0         88        0   135
#> [598,]  81  16      1          9        0   180
#> [599,]  86   3      0          3        1   140
#> [600,]  82   9      0        180        1   134
#> [601,]  84   3      0        180        1   120
#> [602,]  81  13      0        180        0   152
#> [603,]  83   9      0        180        1   149
#> [604,]  87   2      0          5        1   137
#> [605,]  86  12      1          0        1   132
#> [606,]  86   6      1          0        1   140
#> [607,]  84  16      0         70        1   150
#> [608,]  84   3      1          2        0   125
#> [609,]  83  10      1          0        1   190
#> [610,]  88  14      1          3        1   130
#> [611,]  83  13      1         12        0   170
#> [612,]  84   7      1          2        0   148
#> [613,]  84   9      0         92        1   110
#> [614,]  84   3      0        180        1   170
#> [615,]  86   4      0         38        1   122
#> [616,]  82   4      0          4        0   130
#> [617,]  86  13      0        177        0   163
#> [618,]  86   6      0          6        1   117
#> [619,]  86   6      1          1        0   112
#> [620,]  88   4      0          4        0   100
#> [621,]  83  20      1          3        1   150
#> [622,]  83   9      0         65        1   150
#> [623,]  87   2      0        180        1   130
#> [624,]  86   6      0         46        0   173
#> [625,]  88   3      0        115        0   110
#> [626,]  88   2      0        180        1    68
#> [627,]  88   4      0          4        0    86
#> [628,]  89   4      0          4        1   153
#> [629,]  89   5      0        119        1   140
#> [630,]  87   1      0          1        0   170
#> [631,]  84   8      0        180        1   119
#> [632,]  85   8      0          8        1   136
#> [633,]  84   2      0        110        1   174
#> [634,]  84   0      0        180        1   136
#> [635,]  89  10      0         46        1   170
#> [636,]  88   1      0          1        0   135
#> [637,]  86   4      0        180        1   145
#> [638,]  91   8      0          8        0   100
#> [639,]  91  10      0        145        0   135
#> [640,]  88   7      0         24        0   119
#> [641,]  90  11      1         10        1   186
#> [642,]  87   6      0        126        1   168
#> [643,]  86  10      0        180        1   137
#> [644,]  86   9      1          7        0   130
#> [645,]  90   4      1          0        0   121
#> [646,]  91   1      0          1        1    74
#> [647,]  87   5      0         36        1   150
#> [648,]  88   3      1          2        0   159
#> [649,]  92   1      0          1        1   167
#> [650,]  89  12      1          0        1   130
#> [651,]  89   2      0        168        0   118
#> [652,]  91   5      0        169        1   176
#> [653,]  92   7      0          7        1   110
#> [654,]  89   4      0          4        1   159
#> [655,]  90  18      0        180        0   188
#> [656,]  90   1      0          1        1   118
#> [657,]  91   2      0          2        1   116
#> [658,]  93   8      0        179        1   110
#> [659,]  94   8      0          8        1   142
#> [660,]  91   1      0        180        0   158
#> [661,]  90  16      0         16        1   106
#> [662,]  90   3      0         67        0   162
#> [663,]  96   3      0         12        1    97
#> [664,]  94   3      0         26        1   144
#> [665,]  91   7      0          7        0   135
#> [666,]  93   0      1          0        1   122
#> [667,]  92   5      0         69        0   139
#> [668,]  92   2      0          2        0   112
#> [669,]  93   4      0        180        1   135
#> [670,]  96   3      1          0        1   104
#> 
#> $y
#>   [1] 180.0+   5.0+   2.0+   5.0+ 180.0+   5.0+   2.0+ 180.0+ 180.0+ 115.0 
#>  [11] 180.0+ 180.0+   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0+
#>  [21]   3.0  180.0+ 180.0+ 180.0+   2.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0+
#>  [31] 180.0+ 180.0+ 180.0+ 180.0+ 150.0  180.0+ 180.0+ 180.0+ 180.0+   6.0+
#>  [41] 180.0+ 180.0+ 180.0+  73.0    5.0+ 161.0+ 180.0+ 180.0+   5.0+ 180.0+
#>  [51] 180.0+ 180.0+ 177.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 172.0+
#>  [61] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0  179.0+ 180.0+ 179.0+ 180.0+
#>  [71] 180.0+   4.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [81] 180.0+ 180.0+   7.0  180.0+   4.0+ 180.0+ 180.0+ 180.0+  77.0  180.0+
#>  [91] 180.0+   5.0  180.0+ 180.0+ 180.0+ 180.0+   4.0+  85.0  180.0+ 180.0+
#> [101]  99.0   16.0+ 180.0+ 152.0+   7.0+   6.0+ 180.0+ 180.0+ 180.0+  13.0+
#> [111] 174.0+  28.0    6.0+   1.0  180.0+ 180.0+ 180.0+ 180.0+   7.0+ 180.0+
#> [121] 180.0+ 180.0+ 180.0+ 180.0+  16.0   15.0+ 180.0+ 180.0+ 180.0+  12.0+
#> [131]   8.0    2.0  180.0+ 180.0+ 140.0    1.0  165.0  180.0+ 180.0+ 180.0+
#> [141] 180.0+   8.0+ 180.0+ 180.0+   0.5  180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [151] 171.0+  15.0  166.0+   4.0+ 147.0+ 180.0+ 180.0+ 180.0+   5.0+ 180.0+
#> [161]   4.0+ 180.0+   9.0+   1.0  180.0+   2.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [171] 180.0+   2.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [181] 161.0+ 180.0+ 180.0+   3.0    1.0  180.0+ 180.0+  10.0+ 180.0+  45.0 
#> [191]   3.0+ 180.0+ 180.0+ 172.0+ 172.0+ 180.0+   8.0  180.0+  15.0  180.0+
#> [201] 180.0+ 180.0+  77.0   13.0+ 180.0+ 180.0+ 170.0  169.0    7.0    7.0+
#> [211]   6.0  180.0+ 180.0+ 180.0+ 180.0+   3.0+ 180.0+ 180.0+  28.0  180.0+
#> [221] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+ 180.0+ 180.0+   9.0 
#> [231]   7.0+  22.0  180.0+  84.0  180.0+ 180.0+ 180.0+   1.0    1.0  180.0+
#> [241] 180.0+   4.0+   3.0+ 167.0    6.0+ 180.0+ 180.0+ 180.0+  12.0  180.0+
#> [251]  36.0    3.0+  88.0   12.0  180.0+ 180.0+   0.5  180.0+  12.0+ 180.0+
#> [261] 180.0+ 180.0+  12.0  180.0+   9.0    3.0  180.0+ 180.0+ 180.0+   2.0+
#> [271] 180.0+ 180.0+ 180.0+   3.0+   2.0+ 103.0   15.0  180.0+  13.0  179.0+
#> [281]  14.0+   3.0    0.5+ 180.0+ 180.0+ 175.0+ 180.0+ 180.0+   8.0    5.0 
#> [291]  16.0    1.0  180.0+ 180.0+ 123.0+   1.0+ 180.0+ 180.0+  15.0  180.0+
#> [301] 180.0+ 174.0+ 180.0+ 180.0+ 180.0+ 180.0+   8.0+ 175.0  180.0+ 180.0+
#> [311] 180.0+   6.0   19.0+  12.0   99.0  180.0+ 179.0+  11.0+   0.5   18.0 
#> [321] 180.0+ 180.0+ 152.0+   1.0   18.0+ 101.0    4.0    5.0    7.0+ 180.0+
#> [331] 180.0+ 180.0+   1.0  180.0+   6.0  180.0+   0.5  180.0+ 180.0+ 180.0+
#> [341]  14.0+   5.0+ 103.0    3.0+ 180.0+   5.0+ 180.0+   8.0+ 180.0+   2.0+
#> [351] 180.0+ 172.0+   7.0    7.0    8.0+  13.0+ 180.0+  51.0   19.0  180.0+
#> [361]   1.0   60.0   76.0  180.0+ 132.0   10.0+ 180.0+ 180.0+ 162.0    7.0+
#> [371]   9.0  180.0+ 180.0+ 180.0+  12.0  180.0+   2.0  173.0+ 180.0+ 180.0+
#> [381] 180.0+ 180.0+  28.0   16.0+ 180.0+ 180.0+   6.0  180.0+ 180.0+   7.0+
#> [391]   3.0+ 180.0+ 180.0+ 180.0+   2.0    3.0+ 180.0+  20.0  180.0+  20.0 
#> [401]   0.5  180.0+ 180.0+   8.0    4.0+  58.0  180.0+ 180.0+ 180.0+   3.0 
#> [411] 180.0+   2.0  180.0+  14.0+ 180.0+ 180.0+ 180.0+ 179.0+   1.0  180.0+
#> [421]  15.0  180.0+  10.0    1.0  180.0+   4.0+ 180.0+  10.0  104.0+  57.0 
#> [431] 180.0+  11.0    3.0+ 180.0+  12.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [441]  34.0  177.0+   5.0  180.0+   7.0   15.0+  10.0  180.0+ 180.0+ 180.0+
#> [451] 180.0+  17.0+ 174.0+   7.0  180.0+  12.0  180.0+  46.0  180.0+   1.0 
#> [461] 180.0+ 180.0+  12.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+  33.0    5.0 
#> [471] 180.0+   1.0    7.0+   3.0  180.0+ 180.0+ 176.0+  18.0  180.0+  47.0 
#> [481]  11.0  180.0+   7.0  180.0+   8.0+   7.0   32.0   10.0  180.0+ 172.0 
#> [491]  12.0  180.0+   8.0  180.0+ 180.0+   2.0  180.0+  11.0   29.0   24.0 
#> [501]  23.0    6.0  180.0+   3.0+ 180.0+ 180.0+ 180.0+   1.0   11.0    4.0 
#> [511]   4.0   10.0+ 180.0+   3.0+   2.0+ 180.0+   1.0    1.0    3.0  180.0+
#> [521]   6.0  180.0+ 180.0+  71.0    8.0   40.0  161.0  180.0+ 164.0   37.0 
#> [531] 180.0+ 175.0+  22.0   15.0+   5.0+   3.0  171.0+  71.0   20.0+   3.0+
#> [541]  10.0   85.0   10.0    6.0    5.0    1.0  180.0+ 108.0  180.0+ 125.0 
#> [551] 180.0+ 180.0+   4.0    6.0    9.0+ 180.0+ 180.0+ 180.0+ 169.0   70.0 
#> [561]   4.0  180.0+ 180.0+ 180.0+  20.0    8.0+  16.0  180.0+ 177.0+ 180.0+
#> [571] 180.0+ 167.0    3.0+  62.0  180.0+   1.0   38.0  180.0+  90.0  180.0+
#> [581]  89.0  180.0+  71.0   30.0  180.0+ 114.0  180.0+ 154.0  180.0+   4.0+
#> [591]   1.0   16.0+ 180.0+  77.0  180.0+   3.0   88.0  180.0+   3.0  180.0+
#> [601] 180.0+ 180.0+ 180.0+   5.0  180.0+   6.0   70.0  180.0+ 180.0+  14.0 
#> [611]  13.0  180.0+  92.0  180.0+  38.0    4.0  177.0    6.0+   6.0+   4.0+
#> [621]  20.0   65.0  180.0+  46.0  115.0  180.0+   4.0    4.0  119.0    1.0+
#> [631] 180.0+   8.0  110.0  180.0+  46.0    1.0+ 180.0+   8.0  145.0   24.0 
#> [641]  11.0  126.0  180.0+ 180.0+   4.0    1.0   36.0   75.0    1.0  180.0+
#> [651] 168.0  169.0    7.0    4.0  180.0+   1.0+   2.0  179.0+   8.0+ 180.0+
#> [661]  16.0   67.0   12.0   26.0    7.0+   0.5   69.0    2.0  180.0+   3.0 
#> 
#> $weights
#> NULL
#> 
#> $offset
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8430299