Cross-Validated GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::cv.glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.cv.glmnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.cv.glmnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.cv_glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

foldid	untyped	NULL		-

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nfolds	integer	10		\([3, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	TRUE	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	lambda.1se		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

standardize.response	logical	FALSE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.gaussian	character	-	covariance, naive	-

type.logistic	character	Newton	Newton, modified.Newton	-

type.measure	character	deviance	deviance, C	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.coxnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvCVGlmnet

Methods

Public methods

LearnerSurvCVGlmnet$new()

LearnerSurvCVGlmnet$selected_features()

LearnerSurvCVGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvCVGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvCVGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvCVGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.cv_glmnet")
print(learner)
#> <LearnerSurvCVGlmnet:surv.cv_glmnet>: Regularized Generalized Linear Model
#> * Model: -
#> * Parameters: use_pred_offset=TRUE
#> * Packages: mlr3, mlr3proba, mlr3extralearners, glmnet
#> * Predict Types:  [crank], distr, lp
#> * Feature Types: logical, integer, numeric
#> * Properties: offset, selected_features, weights

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#> Measure: Partial Likelihood Deviance 
#> 
#>      Lambda Index Measure     SE Nonzero
#> min 0.00329    45   8.707 0.2233       6
#> 1se 0.04884    16   8.918 0.1855       4
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  33   2      0          2        0   150
#>   [2,]  35   5      1          2        0   172
#>   [3,]  35   2      1          1        1   112
#>   [4,]  37   9      0        180        1   151
#>   [5,]  38   2      0        115        0   150
#>   [6,]  36   1      0        180        1   155
#>   [7,]  35   0      0        180        1   119
#>   [8,]  38  12      1          8        1   120
#>   [9,]  36   5      1          0        1   115
#>  [10,]  33   6      1          1        1   115
#>  [11,]  42   3      1          1        1   130
#>  [12,]  37   1      1          0        1   146
#>  [13,]  42   2      0          2        0   140
#>  [14,]  40   6      0        180        1   138
#>  [15,]  40  11      1         10        1   120
#>  [16,]  42   2      0        180        0   100
#>  [17,]  41   2      1          1        0   166
#>  [18,]  43   4      1          0        1   130
#>  [19,]  42   4      0        180        0   162
#>  [20,]  42  15      1         13        1   125
#>  [21,]  42  12      1         10        1   170
#>  [22,]  43   2      1          1        1   116
#>  [23,]  44   5      1          1        0   170
#>  [24,]  45   3      0        180        1   154
#>  [25,]  44   3      0        180        0   141
#>  [26,]  41  13      1          1        0   140
#>  [27,]  45   6      0        180        1   170
#>  [28,]  41   5      1          4        1   141
#>  [29,]  44   2      1          1        1   150
#>  [30,]  43   2      0        180        1   140
#>  [31,]  45   2      0        180        1   140
#>  [32,]  46  15      0        180        0   120
#>  [33,]  48  15      0        180        1   160
#>  [34,]  45   3      0        150        0   130
#>  [35,]  44   3      1          0        1   180
#>  [36,]  46   7      1          2        0   166
#>  [37,]  43  29      0        180        1   180
#>  [38,]  45   4      1          0        0   124
#>  [39,]  47   6      1          0        1   116
#>  [40,]  44   0      1          0        1    96
#>  [41,]  47   4      1          3        1   160
#>  [42,]  43   3      1          0        1   124
#>  [43,]  49   5      0         73        1   136
#>  [44,]  45   5      0          5        0   141
#>  [45,]  46   2      1          1        1   122
#>  [46,]  44   4      1          0        1   114
#>  [47,]  47   2      0        180        0   108
#>  [48,]  45   5      0        180        1   190
#>  [49,]  46   5      1          3        0   130
#>  [50,]  46   4      0        180        1   121
#>  [51,]  44   2      0        180        0   142
#>  [52,]  46  15      0        180        1   120
#>  [53,]  45   9      1          0        1   145
#>  [54,]  47   3      1          1        1   120
#>  [55,]  48  12      1         11        0   200
#>  [56,]  47   5      1          3        1   130
#>  [57,]  47   9      1          6        0   170
#>  [58,]  46   3      1          0        1   119
#>  [59,]  47  10      0         10        1   140
#>  [60,]  50   1      1          0        1   129
#>  [61,]  50   4      1          1        0   125
#>  [62,]  46   3      1          1        1   140
#>  [63,]  50   7      0        180        1   110
#>  [64,]  49   2      0          2        0   105
#>  [65,]  49  15      1         11        1   160
#>  [66,]  49  23      0        179        1   112
#>  [67,]  52   2      0        180        1   170
#>  [68,]  50   7      1          0        1    92
#>  [69,]  50   4      0          4        1   100
#>  [70,]  51   3      1          2        0   113
#>  [71,]  50   1      1          0        0   150
#>  [72,]  47   8      0        180        0   160
#>  [73,]  47   6      0        180        1   162
#>  [74,]  51   8      0        180        1   140
#>  [75,]  52   2      0        180        0   155
#>  [76,]  46   3      0        180        1   120
#>  [77,]  48   7      1          0        1   110
#>  [78,]  53   8      0         36        1   160
#>  [79,]  48  17      1         10        0   111
#>  [80,]  47   2      1          1        0   110
#>  [81,]  52   4      1          4        0   152
#>  [82,]  49  15      0        180        1   160
#>  [83,]  53   5      0         77        0   159
#>  [84,]  50  10      1          6        0   122
#>  [85,]  53   8      1          7        0   160
#>  [86,]  51  25      1          1        0   202
#>  [87,]  49   5      1          2        1   150
#>  [88,]  52  14      1          7        1   200
#>  [89,]  48  11      1         10        0   120
#>  [90,]  53   4      1          0        1   156
#>  [91,]  51  13      0         99        1   160
#>  [92,]  54   9      1          0        1   138
#>  [93,]  49  16      0         16        0   125
#>  [94,]  55   3      1          1        0   150
#>  [95,]  54  23      1         10        0   131
#>  [96,]  52   7      1          2        0   154
#>  [97,]  55   6      1          2        1   114
#>  [98,]  55   4      1          2        0   150
#>  [99,]  51  13      1         11        0   145
#> [100,]  54   4      1          0        1   121
#> [101,]  52   4      0        180        0   183
#> [102,]  50   3      0        174        1   153
#> [103,]  55  28      1         13        1   160
#> [104,]  49   6      1          0        1   130
#> [105,]  53   9      0          9        1    95
#> [106,]  53   8      1          0        1   130
#> [107,]  50   7      1          0        1   127
#> [108,]  55   1      0        180        0   127
#> [109,]  54   1      0        180        0   162
#> [110,]  54   7      1          0        1   100
#> [111,]  56   2      0        180        0   132
#> [112,]  55   5      1          4        1   120
#> [113,]  54   3      0        180        1   180
#> [114,]  52  16      0         16        0   152
#> [115,]  53  10      1          9        0   172
#> [116,]  52  16      1         14        0   170
#> [117,]  53  15      0         15        1    90
#> [118,]  55   6      0        180        1   100
#> [119,]  54  12      1          0        1   190
#> [120,]  55   2      0        134        1   140
#> [121,]  54   3      0        180        0   128
#> [122,]  56   3      0          8        1   139
#> [123,]  57   3      0          3        0   120
#> [124,]  54   7      1          2        0   129
#> [125,]  52   9      1          3        0   170
#> [126,]  54   2      1          1        1   176
#> [127,]  57   1      0        180        1   156
#> [128,]  56   4      1          0        1   140
#> [129,]  52   2      0        180        0   140
#> [130,]  55  11      1          7        0   104
#> [131,]  52  15      1         14        0   130
#> [132,]  53   3      1          0        1   200
#> [133,]  57  10      0        180        1   170
#> [134,]  58   8      0          8        1   130
#> [135,]  54   5      0        180        1   108
#> [136,]  53  21      1         13        1   130
#> [137,]  58   6      1          0        1    90
#> [138,]  53  15      1         10        1   130
#> [139,]  54  17      1          8        1   227
#> [140,]  54  23      1          8        0   120
#> [141,]  53   4      0        147        1   145
#> [142,]  53   7      1          0        1   120
#> [143,]  57  11      1         10        1   129
#> [144,]  55   3      1          2        0   140
#> [145,]  55   5      0          5        1   131
#> [146,]  54   7      1          0        1   141
#> [147,]  59  15      1         10        0   140
#> [148,]  58   9      1          0        1   180
#> [149,]  58   1      1          1        1   200
#> [150,]  56   7      1          5        1   120
#> [151,]  55   2      0          2        0   106
#> [152,]  59   9      1          1        1   125
#> [153,]  57   1      0        180        0   148
#> [154,]  59   3      0        180        0   120
#> [155,]  58   4      1          0        1   160
#> [156,]  60   5      1          1        0   138
#> [157,]  57   5      0        180        1   130
#> [158,]  58  11      1          9        1   124
#> [159,]  57  10      1          9        0   103
#> [160,]  59   6      1          0        1   140
#> [161,]  59   5      0        180        1   155
#> [162,]  59   4      1          0        1   152
#> [163,]  60   0      1          0        1    80
#> [164,]  59   2      1          1        0   140
#> [165,]  58   8      0        161        1   140
#> [166,]  58  14      1          6        0   190
#> [167,]  61   4      1          3        0   151
#> [168,]  61   9      1          8        0   150
#> [169,]  58   1      0          1        1   100
#> [170,]  57  13      1         10        0   110
#> [171,]  57   2      1          0        1   116
#> [172,]  57   4      1          3        0   138
#> [173,]  57  11      0        180        1   150
#> [174,]  56  14      0         45        0   130
#> [175,]  58  19      1         13        1   140
#> [176,]  56  13      1          6        1   158
#> [177,]  56  18      1         11        1   165
#> [178,]  59   9      1          0        1    80
#> [179,]  55   4      1          3        1   160
#> [180,]  58  11      0        172        1   135
#> [181,]  60  12      1          0        1   114
#> [182,]  55   9      1          7        1   135
#> [183,]  61   4      1          0        1   115
#> [184,]  56   8      1          8        0   120
#> [185,]  61  13      1         12        1   130
#> [186,]  59  11      1          8        1   190
#> [187,]  57   1      0          1        0   126
#> [188,]  59   5      1          2        0   182
#> [189,]  58   5      1          1        1   135
#> [190,]  59  10      0        180        0   160
#> [191,]  61   8      0         77        0   120
#> [192,]  61  13      0         13        0   210
#> [193,]  58   8      1          5        0   152
#> [194,]  62  10      1          0        1   153
#> [195,]  62   7      1          2        1   180
#> [196,]  57   3      1          0        0   100
#> [197,]  61  18      0        170        0   140
#> [198,]  61  28      1          7        0   133
#> [199,]  58   8      1          3        1   150
#> [200,]  61   7      0          7        1   150
#> [201,]  61   6      0          6        0   134
#> [202,]  59  13      1          2        0   198
#> [203,]  57  12      1          9        1   120
#> [204,]  62   4      1          0        0   160
#> [205,]  60  17      1          8        1   140
#> [206,]  58   3      1          0        1   146
#> [207,]  62   4      1          3        0   173
#> [208,]  58   2      0         30        0   202
#> [209,]  59   1      0        180        0   155
#> [210,]  63   6      0         28        1   120
#> [211,]  61   5      0          5        1   110
#> [212,]  57  18      1          9        1    93
#> [213,]  61   5      0          5        1   160
#> [214,]  58  11      1          9        0   179
#> [215,]  57   2      1          1        0   159
#> [216,]  62   1      1          0        1   172
#> [217,]  63   1      0        180        1   130
#> [218,]  61   7      0        180        0   135
#> [219,]  63   4      1          3        0   222
#> [220,]  62   3      0        180        1   105
#> [221,]  63   4      0        180        1   190
#> [222,]  63  15      1         10        1   126
#> [223,]  64   4      0        180        0   130
#> [224,]  63   4      1          1        0   155
#> [225,]  60  18      1         13        0   132
#> [226,]  61   9      1          9        1   150
#> [227,]  58   9      1          9        0   110
#> [228,]  62   7      0          7        0   150
#> [229,]  58   2      0        180        0   127
#> [230,]  59   4      0        180        0   196
#> [231,]  60   7      0          7        0   140
#> [232,]  59   5      1          1        0   148
#> [233,]  65  13      0        180        1   100
#> [234,]  63   1      0          1        0   162
#> [235,]  62   6      0        180        0   170
#> [236,]  61  15      1         13        0   170
#> [237,]  60   3      0          3        0   168
#> [238,]  63  12      1         10        0   200
#> [239,]  61   6      1          1        1   117
#> [240,]  64  12      1         11        0   160
#> [241,]  66   1      1          0        1   120
#> [242,]  63  10      1          0        1   148
#> [243,]  63  14      1          9        0   123
#> [244,]  65  36      1         11        0   140
#> [245,]  63   4      1          3        0   162
#> [246,]  66   3      1          1        0   127
#> [247,]  61  10      1          2        1   194
#> [248,]  64  32      1          9        1   160
#> [249,]  63   7      0        180        0   120
#> [250,]  66   5      1          0        1   110
#> [251,]  65   8      1          0        0   168
#> [252,]  60   6      0        180        0   130
#> [253,]  61  12      1         11        0   154
#> [254,]  64   9      0        180        0   150
#> [255,]  61   4      0        180        1   113
#> [256,]  65   3      0        180        1   190
#> [257,]  63  16      1          7        1   110
#> [258,]  62   3      1          1        1   199
#> [259,]  65   3      1          0        1    80
#> [260,]  63   2      1          1        0   180
#> [261,]  62  13      1         11        0   180
#> [262,]  67  11      0         11        1   100
#> [263,]  66  18      1          5        0   142
#> [264,]  66  16      1         11        1   169
#> [265,]  61  14      1          5        0   140
#> [266,]  63   9      1          8        1   160
#> [267,]  64  19      1          8        1   160
#> [268,]  67   6      0        180        1   170
#> [269,]  64   6      1          0        1   125
#> [270,]  66  13      1          0        0   118
#> [271,]  65   3      0          3        0   105
#> [272,]  66   3      1          0        1   135
#> [273,]  66   6      1          0        1   140
#> [274,]  65   2      1          1        1   170
#> [275,]  68   1      0        180        1   166
#> [276,]  64  10      1          9        1   110
#> [277,]  63   7      1          0        0   162
#> [278,]  63  10      0         16        1   160
#> [279,]  64   1      0          1        1   120
#> [280,]  68  18      0        180        1   260
#> [281,]  65  17      1         14        1   100
#> [282,]  63   8      1          1        1   162
#> [283,]  65  18      1          3        0   120
#> [284,]  63   1      1          0        1   155
#> [285,]  67  11      0         11        0   150
#> [286,]  68  11      0        180        0   160
#> [287,]  66  11      1          0        0   100
#> [288,]  66  15      1         13        1   160
#> [289,]  65  11      1          6        0   130
#> [290,]  69  21      1         10        0   180
#> [291,]  69   6      0        180        1   100
#> [292,]  66   9      1          8        0   130
#> [293,]  63   8      0        180        1   120
#> [294,]  68  14      1         13        1   140
#> [295,]  65   8      1          0        1    90
#> [296,]  69   1      1          0        0   170
#> [297,]  67   1      0        180        1   160
#> [298,]  65   1      1          0        0   133
#> [299,]  67   7      1          4        1   130
#> [300,]  63   2      1          0        0    99
#> [301,]  67   2      0        180        0   184
#> [302,]  65   6      0          6        0    80
#> [303,]  65  10      1          1        1   148
#> [304,]  69   6      0         99        1   140
#> [305,]  65   4      1          1        0   130
#> [306,]  66   4      0        180        1   130
#> [307,]  70  15      1         12        1   132
#> [308,]  64  11      0         11        0   125
#> [309,]  64   4      0        180        1   140
#> [310,]  64   0      1          0        1   118
#> [311,]  66   4      0        180        0   177
#> [312,]  69   4      1          3        1   150
#> [313,]  69  17      1         10        0   140
#> [314,]  69   8      0         93        0   140
#> [315,]  66   6      0        180        0   140
#> [316,]  65   1      0          1        1   120
#> [317,]  68  18      1          0        1   160
#> [318,]  65   6      0        101        1   115
#> [319,]  68   4      0          4        1   190
#> [320,]  70   7      1          0        1   190
#> [321,]  68   7      0        150        0   210
#> [322,]  71  20      1          0        1   160
#> [323,]  66   9      1          3        1   151
#> [324,]  69   8      0        180        1   153
#> [325,]  70  14      0        171        0   166
#> [326,]  66   4      0        180        0   130
#> [327,]  67  10      1          9        0   200
#> [328,]  67   6      1          4        0   130
#> [329,]  68  18      1         14        1   170
#> [330,]  69   0      0          0        1   148
#> [331,]  65   2      0        180        0   130
#> [332,]  69   3      1          2        0   151
#> [333,]  67  14      1         13        0   130
#> [334,]  65  14      1         13        1   150
#> [335,]  71   7      0          7        0   230
#> [336,]  66   2      0          2        1   228
#> [337,]  71   6      0         45        1   158
#> [338,]  71   3      0        103        0   133
#> [339,]  69   3      0          3        1   130
#> [340,]  67   1      0         36        1   104
#> [341,]  67   5      0          5        0   130
#> [342,]  68   6      0        180        0   145
#> [343,]  69   6      1          4        1   174
#> [344,]  72   7      0          7        1   110
#> [345,]  69   8      1          7        1   108
#> [346,]  67   3      0        180        0   110
#> [347,]  66   2      1          1        0   123
#> [348,]  66   2      0        180        0   130
#> [349,]  67   7      1          4        0   122
#> [350,]  69   4      1          3        0   132
#> [351,]  69   8      1          2        0   121
#> [352,]  67  13      1          9        0   130
#> [353,]  70   9      0        180        1   142
#> [354,]  72   5      1          4        0   170
#> [355,]  67  22      1          1        1   140
#> [356,]  69   1      0          1        1   110
#> [357,]  67   1      0          1        1    60
#> [358,]  67   4      0         60        1   136
#> [359,]  67   8      1          0        1   130
#> [360,]  72  13      1         11        1   195
#> [361,]  68  10      1          8        1   160
#> [362,]  70  35      1          0        1   105
#> [363,]  72  30      1          0        1   145
#> [364,]  68   7      1          2        0   135
#> [365,]  71   6      0          9        0   120
#> [366,]  67   8      0        180        1   170
#> [367,]  67   5      1          0        1   147
#> [368,]  67   9      0        180        0   158
#> [369,]  73  13      0        152        1   130
#> [370,]  70   5      0        180        0   150
#> [371,]  72   2      0          2        1   100
#> [372,]  67   4      1          1        0   134
#> [373,]  71   1      0        173        1   188
#> [374,]  69   3      0        180        0   220
#> [375,]  72   5      0         28        0   120
#> [376,]  71   5      0        180        0   191
#> [377,]  73   6      0        180        1   117
#> [378,]  69  16      1         10        1   140
#> [379,]  73   6      1          0        1   270
#> [380,]  72   8      1          1        1   150
#> [381,]  73   7      0          7        1   140
#> [382,]  68  15      1         13        1   130
#> [383,]  70   3      0          3        1   159
#> [384,]  70  13      1          9        0   100
#> [385,]  73   0      0        180        1   161
#> [386,]  73   4      0        180        1   154
#> [387,]  69   2      1          0        1   110
#> [388,]  71   3      1          1        0   150
#> [389,]  71  15      1         11        0   165
#> [390,]  74  20      0         20        1   180
#> [391,]  68   9      0        180        1   120
#> [392,]  71  20      1         10        0   140
#> [393,]  73   3      1          0        1   136
#> [394,]  70   5      1          0        1   190
#> [395,]  71  17      1         11        0   160
#> [396,]  71   3      1          2        1   190
#> [397,]  69  12      1          1        1   149
#> [398,]  72   5      1          3        1   160
#> [399,]  70   3      0        180        1   154
#> [400,]  73   6      0        180        0   110
#> [401,]  72  15      1          0        1   150
#> [402,]  72   8      1          0        1   140
#> [403,]  74   3      0          3        1   150
#> [404,]  71  13      1          8        0   121
#> [405,]  69   2      1          1        1    80
#> [406,]  71  14      1         13        1   170
#> [407,]  72  10      1          8        1   153
#> [408,]  69   7      0        180        1   144
#> [409,]  72  15      1         13        0   156
#> [410,]  70   8      0          8        0   120
#> [411,]  71  10      1          9        1   120
#> [412,]  75   1      0          1        0   133
#> [413,]  75   2      1          1        0   145
#> [414,]  73  10      1          9        1   146
#> [415,]  72  10      1          9        1   160
#> [416,]  73  10      1         10        1   120
#> [417,]  73   1      0          1        1    80
#> [418,]  75   9      1          7        0   140
#> [419,]  75  13      1          1        1   130
#> [420,]  71  11      1          8        0   110
#> [421,]  71   4      0          4        0   134
#> [422,]  73  10      1          8        0   120
#> [423,]  70   7      1          4        0   184
#> [424,]  72   1      1          1        0   168
#> [425,]  72   7      0         57        1   145
#> [426,]  72  11      0         11        1   140
#> [427,]  70   3      0          3        0   150
#> [428,]  76  25      1         12        1   170
#> [429,]  73  12      1         12        1   140
#> [430,]  72   2      0        180        0   120
#> [431,]  75   1      0        180        1   140
#> [432,]  72   4      1          0        1   197
#> [433,]  71   3      1          0        0   144
#> [434,]  73   5      0        180        0   126
#> [435,]  72   5      0        180        0   154
#> [436,]  75   3      1          1        0   180
#> [437,]  72   7      1          2        0   142
#> [438,]  73  15      0         15        1   160
#> [439,]  71  16      0        180        0   140
#> [440,]  74   7      0        180        1   150
#> [441,]  74   3      0          3        1   128
#> [442,]  76   1      0        180        0   114
#> [443,]  76   8      1          0        1   141
#> [444,]  73   6      0          6        1   114
#> [445,]  74   2      0        180        0   190
#> [446,]  72   4      0         85        1   120
#> [447,]  76  17      1          0        1   200
#> [448,]  76  13      1         10        0   110
#> [449,]  75   4      1          0        1   122
#> [450,]  75   7      0          7        0   190
#> [451,]  75   0      0          0        1   130
#> [452,]  73  13      1         11        0   195
#> [453,]  75  12      0         12        1   160
#> [454,]  74   8      1          0        1   105
#> [455,]  75   4      1          2        1   188
#> [456,]  74   6      0        180        0   160
#> [457,]  76   4      0          4        1   155
#> [458,]  74   2      0        180        0   111
#> [459,]  73   1      0         52        1   105
#> [460,]  76  44      1         10        0   105
#> [461,]  76   5      0        180        0   185
#> [462,]  74  10      1          0        1   135
#> [463,]  76   5      1          0        1   167
#> [464,]  75   9      0        180        1   140
#> [465,]  77   5      1          0        0   123
#> [466,]  77  12      1          9        1   100
#> [467,]  73  10      1          9        0   146
#> [468,]  77   1      1          0        1    90
#> [469,]  76  12      1         11        1   120
#> [470,]  78   5      1          0        1   170
#> [471,]  75   3      1          1        1   171
#> [472,]  75   6      0        180        0   150
#> [473,]  76  29      0         47        0    90
#> [474,]  73   8      1          1        1   162
#> [475,]  78   7      0          7        1   133
#> [476,]  74  15      0        180        1   172
#> [477,]  78   8      1          6        1   110
#> [478,]  74   7      0          7        0   161
#> [479,]  76  13      1          1        1   170
#> [480,]  78  32      1          9        1   198
#> [481,]  79   6      0        180        0   170
#> [482,]  78  13      1          5        0   130
#> [483,]  75   5      0        119        1   150
#> [484,]  78  15      0        180        1   270
#> [485,]  74  10      1          8        0   135
#> [486,]  76   1      0          1        1    83
#> [487,]  78  12      1          9        0   150
#> [488,]  77   2      1          0        1   143
#> [489,]  76  11      1          0        0   120
#> [490,]  75   3      0          3        0     0
#> [491,]  76   7      0         29        1   150
#> [492,]  77  24      0         24        1   160
#> [493,]  80   6      0          6        1   150
#> [494,]  76   3      1          0        1   140
#> [495,]  78  11      1          1        1   140
#> [496,]  79  11      0        180        0   160
#> [497,]  79   2      1          0        1   121
#> [498,]  76   4      0          4        1   160
#> [499,]  79   4      0          4        1   125
#> [500,]  76  10      1          8        0   180
#> [501,]  80   3      1          0        1   120
#> [502,]  76   1      0          1        1   140
#> [503,]  77  31      1          3        1   161
#> [504,]  76   1      0          1        1    90
#> [505,]  78   7      1          0        1   110
#> [506,]  79   3      0          3        0   120
#> [507,]  77   7      0        180        1   170
#> [508,]  77   6      0          6        1   144
#> [509,]  79   4      1          0        1   120
#> [510,]  81   1      0        180        0   120
#> [511,]  82   5      0          8        1   120
#> [512,]  78   4      0         59        1   112
#> [513,]  79  10      0         10        1   120
#> [514,]  79  28      0        164        0   100
#> [515,]  80   9      0        118        1   186
#> [516,]  80   6      0        173        1   160
#> [517,]  79   1      0         37        1   140
#> [518,]  81   3      0        180        0   184
#> [519,]  81   2      0        175        0   172
#> [520,]  78   7      0          7        1   147
#> [521,]  77  13      1          0        1   190
#> [522,]  78  15      0         15        0   165
#> [523,]  80   5      1          1        1   108
#> [524,]  78   4      0        180        0   175
#> [525,]  79   3      0          3        1   101
#> [526,]  78  26      1          5        0   194
#> [527,]  78  20      1          0        1   109
#> [528,]  80   1      0          1        0   100
#> [529,]  78   3      1          1        1   152
#> [530,]  77  10      1          8        1   130
#> [531,]  82   3      1          1        1   144
#> [532,]  77   5      0         85        0   188
#> [533,]  80   2      1          1        0   168
#> [534,]  78   2      0        180        0   148
#> [535,]  82   1      1          0        1    82
#> [536,]  81   1      0        108        0   129
#> [537,]  78  12      0        180        0   134
#> [538,]  79   1      0        125        0   193
#> [539,]  82  21      1          2        0   155
#> [540,]  84  22      1         10        0   180
#> [541,]  79   4      0          4        1   121
#> [542,]  80   6      0          6        1   110
#> [543,]  83   9      1          5        1   170
#> [544,]  83   5      0        180        0   148
#> [545,]  79   7      1          6        0   130
#> [546,]  83   4      0        103        0    97
#> [547,]  81   5      0        177        0    41
#> [548,]  78  23      1         10        1   145
#> [549,]  78   9      1          4        1   120
#> [550,]  79   1      0        180        1   170
#> [551,]  80   7      1          0        1   146
#> [552,]  84   5      1          1        1    85
#> [553,]  83   8      0          8        0   115
#> [554,]  81  16      0         16        1   110
#> [555,]  80   6      1          0        1   150
#> [556,]  80  11      1          8        0   110
#> [557,]  81   8      0        180        0   146
#> [558,]  80   8      1          7        0   160
#> [559,]  79   7      0        177        0   197
#> [560,]  85   4      0        180        0    90
#> [561,]  83   2      0          2        1   155
#> [562,]  84   4      0        167        0   198
#> [563,]  82  23      1          0        0   110
#> [564,]  84   5      0        180        1   203
#> [565,]  84   4      0          4        1    85
#> [566,]  84   1      0         38        1   205
#> [567,]  83   3      0        180        0   174
#> [568,]  81   4      0         90        1   138
#> [569,]  79   9      1          8        0   150
#> [570,]  85   3      1          2        1   160
#> [571,]  80  13      1          8        1   140
#> [572,]  79   4      0          4        1    60
#> [573,]  80   6      0         71        1   189
#> [574,]  83   1      0          1        1   100
#> [575,]  82  19      0         19        0   120
#> [576,]  80  30      1         13        0   220
#> [577,]  83   9      0        180        0   198
#> [578,]  79  14      1          0        0   110
#> [579,]  83   3      0        114        0    98
#> [580,]  81  14      1         12        1   128
#> [581,]  83   2      0        154        0   130
#> [582,]  82   0      0          2        1   100
#> [583,]  83   1      0        180        0   160
#> [584,]  81   4      0          4        0   175
#> [585,]  84  15      1         13        1   110
#> [586,]  81   1      0          1        1   145
#> [587,]  81  12      0         12        1   163
#> [588,]  81   4      0          4        0   160
#> [589,]  86  12      0        180        1   120
#> [590,]  83  12      1          2        1   170
#> [591,]  81  19      1         14        0   120
#> [592,]  82   3      1          2        0   130
#> [593,]  82  15      1          0        0   183
#> [594,]  80   2      0         88        0   135
#> [595,]  86   8      0          8        1   132
#> [596,]  81  16      1          9        0   180
#> [597,]  84   6      0        165        0   145
#> [598,]  82   9      0        180        1   134
#> [599,]  81  13      0        180        0   152
#> [600,]  81   2      1          0        1   118
#> [601,]  81   4      0        180        0   160
#> [602,]  83   9      0        180        1   149
#> [603,]  82   1      0        180        1   193
#> [604,]  86  12      1          0        1   132
#> [605,]  84  16      0         70        1   150
#> [606,]  83  19      0         43        0   150
#> [607,]  84   3      1          2        0   125
#> [608,]  83  10      1          0        1   190
#> [609,]  86   2      0        180        1   169
#> [610,]  88  14      1          3        1   130
#> [611,]  84   3      0          3        1   121
#> [612,]  83  13      1         12        0   170
#> [613,]  87   2      0        180        0   113
#> [614,]  84   9      0         92        1   110
#> [615,]  86   4      0         38        1   122
#> [616,]  82   4      0          4        0   130
#> [617,]  86  13      0        177        0   163
#> [618,]  85   3      0          3        1   113
#> [619,]  84  13      0         62        1   100
#> [620,]  85  22      0         22        1   184
#> [621,]  86   9      1          7        1   142
#> [622,]  86   6      0         46        0   173
#> [623,]  88   3      0        115        0   110
#> [624,]  88   2      0        180        1    68
#> [625,]  87   8      0          8        1   157
#> [626,]  86  15      1          8        1   109
#> [627,]  88   4      0          4        0    86
#> [628,]  89   4      0          4        1   153
#> [629,]  87   6      0        180        1   110
#> [630,]  87   1      0          1        0   170
#> [631,]  84   2      0        110        1   174
#> [632,]  84   0      0        180        1   136
#> [633,]  89  10      0         46        1   170
#> [634,]  88   1      0          1        0   135
#> [635,]  86   4      0        180        1   145
#> [636,]  87   6      1          0        0   125
#> [637,]  88   7      0         24        0   119
#> [638,]  90  11      1         10        1   186
#> [639,]  86  10      0        180        1   137
#> [640,]  86   9      1          7        0   130
#> [641,]  90   5      1          0        1   125
#> [642,]  88   3      1          2        0   159
#> [643,]  89   3      1          1        1   160
#> [644,]  92   1      0          1        1   167
#> [645,]  91   3      0         33        1   137
#> [646,]  88   5      0        158        0   100
#> [647,]  87   7      0         74        1   105
#> [648,]  89  12      1          0        1   130
#> [649,]  91   5      0        169        1   176
#> [650,]  89  52      0         52        1   130
#> [651,]  92   7      0          7        1   110
#> [652,]  89   4      0          4        1   159
#> [653,]  91   0      0          0        0     0
#> [654,]  90  18      0        180        0   188
#> [655,]  94   6      0         50        0    78
#> [656,]  90   1      0          1        1   118
#> [657,]  91   2      0          2        1   116
#> [658,]  94   8      0          8        1   142
#> [659,]  91   1      0        180        0   158
#> [660,]  90  16      0         16        1   106
#> [661,]  90   3      0         67        0   162
#> [662,]  95   8      1          5        1   150
#> [663,]  94   3      0         26        1   144
#> [664,]  91  12      0         53        1   212
#> [665,]  91   7      0          7        0   135
#> [666,]  93   0      1          0        1   122
#> [667,]  92   5      0         69        0   139
#> [668,]  93   4      0        180        1   135
#> [669,]  96   3      1          0        1   104
#> [670,]  96  15      1          0        1   140
#> 
#> $y
#>   [1]   2.0+   5.0+   2.0+ 180.0+ 115.0  180.0+ 180.0+  12.0    5.0+ 180.0+
#>  [11] 180.0+ 180.0+   2.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [21] 180.0+   2.0+ 155.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0+ 180.0+ 180.0+
#>  [31] 180.0+ 180.0+ 180.0+ 150.0  180.0+ 180.0+ 180.0+ 180.0+   6.0+ 180.0+
#>  [41] 180.0+ 180.0+  73.0    5.0+ 161.0+ 180.0+ 180.0+ 180.0+   5.0+ 180.0+
#>  [51] 180.0+ 180.0+ 177.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  10.0+ 172.0+
#>  [61] 180.0+ 180.0+ 180.0+   2.0  179.0+ 179.0+ 180.0+ 180.0+   4.0+ 180.0+
#>  [71] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   7.0   36.0   88.0+ 180.0+
#>  [81]   4.0+ 180.0+  77.0  180.0+ 180.0+ 180.0+ 180.0+  85.0  180.0+ 166.0+
#>  [91]  99.0  180.0+  16.0+ 180.0+ 152.0+   7.0+   6.0+ 180.0+  13.0+ 180.0+
#> [101] 180.0+ 174.0+  28.0    6.0+   9.0+ 180.0+ 180.0+ 180.0+ 180.0+   7.0+
#> [111] 180.0+ 180.0+ 180.0+  16.0+ 180.0+  16.0   15.0+ 180.0+  12.0+ 134.0+
#> [121] 180.0+   8.0    3.0+ 180.0+ 180.0+ 180.0+ 180.0+ 165.0  180.0+ 180.0+
#> [131] 180.0+ 180.0+ 180.0+   8.0+ 180.0+ 180.0+ 180.0+ 180.0+ 171.0+ 180.0+
#> [141] 147.0+ 180.0+ 180.0+ 180.0+   5.0+ 180.0+ 180.0+   9.0+   1.0  180.0+
#> [151]   2.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  64.0 
#> [161] 180.0+ 180.0+   0.5  180.0+ 161.0+ 171.0+ 180.0+ 180.0+   1.0  180.0+
#> [171] 180.0+ 180.0+ 180.0+  45.0   19.0  180.0+ 180.0+   9.0+ 180.0+ 172.0+
#> [181] 172.0+  24.0  180.0+   8.0  180.0+ 180.0+   1.0+ 180.0+ 180.0+ 180.0+
#> [191]  77.0   13.0+   8.0+ 180.0+ 180.0+ 180.0+ 170.0   94.0  180.0+   7.0 
#> [201]   6.0  180.0+ 180.0+ 180.0+ 180.0+   3.0+ 180.0+  30.0  180.0+  28.0 
#> [211]   5.0   18.0    5.0+ 180.0+ 180.0+   1.0  180.0+ 180.0+ 180.0+ 180.0+
#> [221] 180.0+ 180.0+ 180.0+   4.0+ 180.0+ 180.0+   9.0    7.0+ 180.0+ 180.0+
#> [231]   7.0+ 180.0+ 180.0+   1.0  180.0+ 180.0+   3.0+ 180.0+ 180.0+  12.0 
#> [241] 180.0+ 180.0+  14.0+  36.0  180.0+   3.0+  88.0  180.0+ 180.0+ 180.0+
#> [251] 180.0+ 180.0+  12.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   3.0  180.0+
#> [261] 180.0+  11.0+  18.0+ 180.0+ 180.0+ 180.0+ 103.0  180.0+ 180.0+ 166.0+
#> [271]   3.0    3.0+ 180.0+ 175.0+ 180.0+ 180.0+   7.0+  16.0    1.0  180.0+
#> [281] 180.0+ 180.0+ 123.0+   1.0+  11.0+ 180.0+ 180.0+ 180.0+ 180.0+ 174.0+
#> [291] 180.0+ 180.0+ 180.0+ 180.0+   8.0+ 175.0  180.0+ 180.0+ 180.0+ 180.0+
#> [301] 180.0+   6.0  180.0+  99.0  180.0+ 180.0+ 180.0+  11.0+ 180.0+   0.5 
#> [311] 180.0+ 152.0+ 180.0+  93.0  180.0+   1.0   18.0+ 101.0    4.0    7.0+
#> [321] 150.0  180.0+ 180.0+ 180.0+ 171.0  180.0+ 174.0+   6.0  180.0+   0.5 
#> [331] 180.0+ 180.0+ 180.0+  14.0+   7.0+   2.0   45.0  103.0    3.0+  36.0 
#> [341]   5.0+ 180.0+  97.0    7.0    8.0+ 180.0+   2.0+ 180.0+   7.0  180.0+
#> [351]   8.0+  13.0+ 180.0+ 180.0+  51.0    1.0    1.0   60.0  180.0+ 132.0 
#> [361]  10.0+ 180.0+ 162.0    7.0+   9.0  180.0+ 180.0+ 180.0+ 152.0  180.0+
#> [371]   2.0   76.0  173.0+ 180.0+  28.0  180.0+ 180.0+  16.0+   6.0  180.0+
#> [381]   7.0+  15.0    3.0+  13.0+ 180.0+ 180.0+   2.0    3.0+ 180.0+  20.0 
#> [391] 180.0+  20.0  180.0+ 180.0+ 180.0+   3.0   12.0  180.0+ 180.0+ 180.0+
#> [401] 180.0+ 180.0+   3.0  175.0    2.0   14.0+  10.0+ 180.0+ 180.0+   8.0+
#> [411] 179.0+   1.0  180.0+ 180.0+ 159.0   15.0    1.0  180.0+  13.0  180.0+
#> [421]   4.0+  10.0  104.0+   1.0   57.0   11.0    3.0+ 180.0+  12.0  180.0+
#> [431] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   7.0   15.0+ 180.0+ 180.0+
#> [441]   3.0  180.0+ 180.0+   6.0  180.0+  85.0   17.0+ 174.0+   4.0    7.0 
#> [451]   0.5  180.0+  12.0  180.0+  46.0  180.0+   4.0  180.0+  52.0  180.0+
#> [461] 180.0+ 180.0+ 180.0+ 180.0+   5.0  180.0+ 180.0+   1.0   12.0  180.0+
#> [471]   3.0  180.0+  47.0  180.0+   7.0  180.0+   8.0+   7.0  180.0+  32.0 
#> [481] 180.0+ 172.0  119.0  180.0+ 180.0+   1.0  180.0+   2.0   11.0    3.0 
#> [491]  29.0   24.0    6.0    3.0+ 180.0+ 180.0+ 180.0+   4.0    4.0   10.0+
#> [501]   3.0+   1.0  171.0    1.0   43.0    3.0  180.0+   6.0  138.0  180.0+
#> [511]   8.0   59.0   10.0+ 164.0  118.0  173.0   37.0  180.0+ 175.0+   7.0+
#> [521]  22.0   15.0+   5.0+ 180.0+   3.0  171.0+  20.0+   1.0    3.0+  10.0 
#> [531] 180.0+  85.0   10.0  180.0+   1.0  108.0  180.0+ 125.0  180.0+ 180.0+
#> [541]   4.0    6.0    9.0+ 180.0+ 180.0+ 103.0  177.0+  70.0  180.0+ 180.0+
#> [551]   7.0+ 180.0+   8.0+  16.0  180.0+ 180.0+ 180.0+ 180.0+ 177.0+ 180.0+
#> [561]   2.0  167.0   62.0  180.0+   4.0   38.0  180.0+  90.0  180.0+ 180.0+
#> [571] 180.0+   4.0   71.0    1.0   19.0   30.0  180.0+ 180.0+ 114.0  180.0+
#> [581] 154.0    2.0  180.0+   4.0+ 180.0+   1.0   12.0    4.0+ 180.0+  77.0 
#> [591] 180.0+   3.0   83.0   88.0    8.0  180.0+ 165.0  180.0+ 180.0+ 180.0+
#> [601] 180.0+ 180.0+ 180.0+ 180.0+  70.0   43.0  180.0+ 180.0+ 180.0+  14.0 
#> [611]   3.0   13.0  180.0+  92.0   38.0    4.0  177.0    3.0+  62.0   22.0 
#> [621]  11.0   46.0  115.0  180.0+   8.0+ 180.0+   4.0    4.0  180.0+   1.0+
#> [631] 110.0  180.0+  46.0    1.0+ 180.0+  25.0   24.0   11.0  180.0+ 180.0+
#> [641]  89.0   75.0    3.0+   1.0   33.0  158.0   74.0  180.0+ 169.0   52.0 
#> [651]   7.0    4.0    0.5  180.0+  50.0    1.0+   2.0    8.0+ 180.0+  16.0 
#> [661]  67.0    8.0   26.0   53.0    7.0+   0.5   69.0  180.0+   3.0   15.0+
#> 
#> $weights
#> NULL
#> 
#> $offset
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8272911