Cross-Validated GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::cv.glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.cv.glmnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.cv.glmnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.cv_glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

foldid	untyped	NULL		-

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nfolds	integer	10		\([3, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	TRUE	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	lambda.1se		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

standardize.response	logical	FALSE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.gaussian	character	-	covariance, naive	-

type.logistic	character	Newton	Newton, modified.Newton	-

type.measure	character	deviance	deviance, C	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.coxnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvCVGlmnet

Methods

Public methods

LearnerSurvCVGlmnet$new()

LearnerSurvCVGlmnet$selected_features()

LearnerSurvCVGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvCVGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvCVGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvCVGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("surv.cv_glmnet")
print(learner)
#> 
#> ── <LearnerSurvCVGlmnet> (surv.cv_glmnet): Regularized Generalized Linear Model 
#> • Model: -
#> • Parameters: use_pred_offset=TRUE
#> • Packages: mlr3, mlr3proba, mlr3extralearners, and glmnet
#> • Predict Types: [crank], distr, and lp
#> • Feature Types: logical, integer, and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: offset, selected_features, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("grace")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#> Measure: Partial Likelihood Deviance 
#> 
#>      Lambda Index Measure     SE Nonzero
#> min 0.00366    44   2.851 0.1915       6
#> 1se 0.07883    11   3.011 0.1850       3
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  32   5      1          0        1   121
#>   [2,]  35  10      1          9        0   106
#>   [3,]  34   5      0          5        0   120
#>   [4,]  35   2      0        180        0   121
#>   [5,]  37   9      0        180        1   151
#>   [6,]  35   0      0        180        1   119
#>   [7,]  38  12      1          8        1   120
#>   [8,]  36   5      1          0        1   115
#>   [9,]  33   6      1          1        1   115
#>  [10,]  38  16      1         10        0   160
#>  [11,]  40  12      1          9        0   153
#>  [12,]  42   3      1          1        1   130
#>  [13,]  37   1      1          0        1   146
#>  [14,]  40   2      1          1        1   148
#>  [15,]  42   2      0          2        0   140
#>  [16,]  40   6      0        180        1   138
#>  [17,]  40  11      1         10        1   120
#>  [18,]  42   2      0        180        0   100
#>  [19,]  41   2      1          1        0   166
#>  [20,]  43   4      1          0        1   130
#>  [21,]  42   4      0        180        0   162
#>  [22,]  42  15      1         13        1   125
#>  [23,]  42  12      1         10        1   170
#>  [24,]  42   2      0        180        1   124
#>  [25,]  44   5      1          1        0   170
#>  [26,]  41  10      1          8        0   150
#>  [27,]  44   3      0        180        0   141
#>  [28,]  41  13      1          1        0   140
#>  [29,]  45   6      0        180        1   170
#>  [30,]  41   5      1          4        1   141
#>  [31,]  44   2      1          1        1   150
#>  [32,]  43   2      0        180        1   140
#>  [33,]  45   2      0        180        1   140
#>  [34,]  47   4      1          3        0   118
#>  [35,]  48  15      0        180        1   160
#>  [36,]  45   3      0        150        0   130
#>  [37,]  46   7      1          2        0   166
#>  [38,]  43  29      0        180        1   180
#>  [39,]  47   6      1          0        1   116
#>  [40,]  46  13      1         10        0   100
#>  [41,]  44   0      1          0        1    96
#>  [42,]  43   3      1          0        1   124
#>  [43,]  49   5      0         73        1   136
#>  [44,]  45   5      0          5        0   141
#>  [45,]  46   6      1          0        1   100
#>  [46,]  47   2      0        180        0   108
#>  [47,]  44   9      1          8        1   135
#>  [48,]  45   5      0        180        1   190
#>  [49,]  46   5      1          3        0   130
#>  [50,]  46   4      0        180        1   121
#>  [51,]  44   2      0        180        0   142
#>  [52,]  46  15      0        180        1   120
#>  [53,]  45   9      1          0        1   145
#>  [54,]  48   3      0        180        0   154
#>  [55,]  48  12      1         11        0   200
#>  [56,]  47   5      1          3        1   130
#>  [57,]  47   9      1          6        0   170
#>  [58,]  46   3      1          0        1   119
#>  [59,]  49   4      0        180        0   117
#>  [60,]  47   7      0        180        0   145
#>  [61,]  50   4      1          1        0   125
#>  [62,]  46   3      1          1        1   140
#>  [63,]  46   9      1          9        1   122
#>  [64,]  50   7      0        180        1   110
#>  [65,]  49   2      0          2        0   105
#>  [66,]  51   1      0          1        1   145
#>  [67,]  49  23      0        179        1   112
#>  [68,]  51   3      1          2        0   113
#>  [69,]  50   1      1          0        0   150
#>  [70,]  50   9      0        180        0   130
#>  [71,]  47   8      0        180        0   160
#>  [72,]  51   8      0        180        1   140
#>  [73,]  46   3      0        180        1   120
#>  [74,]  46   1      1          1        0   145
#>  [75,]  50   4      1          1        0   150
#>  [76,]  48   7      1          0        1   110
#>  [77,]  53   8      0         36        1   160
#>  [78,]  48  17      1         10        0   111
#>  [79,]  47   2      1          1        0   110
#>  [80,]  49   9      1          3        0   102
#>  [81,]  49  15      0        180        1   160
#>  [82,]  54  17      1         12        1   102
#>  [83,]  53   5      0         77        0   159
#>  [84,]  53   7      1          0        0   199
#>  [85,]  54   6      1          3        0   129
#>  [86,]  51   3      1          1        0   140
#>  [87,]  50   2      0          5        1   106
#>  [88,]  50  10      1          6        0   122
#>  [89,]  50  14      1         13        0   170
#>  [90,]  48   3      1          2        0   150
#>  [91,]  51  25      1          1        0   202
#>  [92,]  53   4      0          4        0   140
#>  [93,]  48  11      1         10        0   120
#>  [94,]  53   4      1          0        1   156
#>  [95,]  49  16      0         16        0   125
#>  [96,]  54  23      1         10        0   131
#>  [97,]  52   7      1          2        0   154
#>  [98,]  54   9      1          1        0   130
#>  [99,]  52   4      0        180        1   180
#> [100,]  51  13      1         11        0   145
#> [101,]  52   4      0        180        0   183
#> [102,]  50   3      0        174        1   153
#> [103,]  55  28      1         13        1   160
#> [104,]  49   6      1          0        1   130
#> [105,]  49   1      0          1        1   110
#> [106,]  50   7      1          1        0   156
#> [107,]  53   9      0          9        1    95
#> [108,]  53   8      1          0        1   130
#> [109,]  50   7      1          0        1   127
#> [110,]  56   4      1          1        1   130
#> [111,]  52   5      0        175        1   117
#> [112,]  55   2      0          2        0   145
#> [113,]  54   1      0        180        0   162
#> [114,]  54   7      1          0        1   100
#> [115,]  56   3      0        180        1   193
#> [116,]  55   5      1          4        1   120
#> [117,]  54   3      0        180        1   180
#> [118,]  55   6      0        180        0   170
#> [119,]  52  16      0         16        0   152
#> [120,]  53  10      1          9        0   172
#> [121,]  53  15      0         15        1    90
#> [122,]  55   6      0        180        1   100
#> [123,]  55   6      1          5        1   138
#> [124,]  54   3      0        180        0   128
#> [125,]  56   3      0          8        1   139
#> [126,]  55   1      0          2        0   130
#> [127,]  57   3      0          3        0   120
#> [128,]  54   7      1          2        0   129
#> [129,]  54   2      1          1        1   176
#> [130,]  57   1      0        180        1   156
#> [131,]  57   1      0          1        1   100
#> [132,]  56   4      1          0        1   140
#> [133,]  52   2      0        180        0   140
#> [134,]  55  11      1          7        0   104
#> [135,]  53   3      1          0        1   200
#> [136,]  57  10      0        180        1   170
#> [137,]  54   5      0        180        1   108
#> [138,]  55   3      1          1        1   156
#> [139,]  53  21      1         13        1   130
#> [140,]  57   4      0        180        1   119
#> [141,]  58   6      1          0        1    90
#> [142,]  53  15      1         10        1   130
#> [143,]  54  17      1          8        1   227
#> [144,]  55   9      1          2        1   147
#> [145,]  56   5      0          5        1   150
#> [146,]  53   4      0        147        1   145
#> [147,]  53   7      1          0        1   120
#> [148,]  55   3      1          2        0   140
#> [149,]  55   5      0          5        1   131
#> [150,]  54   7      1          0        1   141
#> [151,]  56   4      0          4        0   164
#> [152,]  59  15      1         10        0   140
#> [153,]  58   9      1          0        1   180
#> [154,]  58   1      1          1        1   200
#> [155,]  55   5      1          0        0   140
#> [156,]  56   7      1          5        1   120
#> [157,]  55   2      0          2        0   106
#> [158,]  57   1      0        180        0   148
#> [159,]  59   3      0        180        0   120
#> [160,]  58   4      1          0        1   160
#> [161,]  57   2      0          2        1   120
#> [162,]  60   5      1          1        0   138
#> [163,]  57   5      0        180        1   130
#> [164,]  58  11      1          9        1   124
#> [165,]  55   5      1          0        1   160
#> [166,]  59   6      1          0        1   140
#> [167,]  59   4      1          0        1   152
#> [168,]  61   9      0          9        1   160
#> [169,]  58   4      1          3        0   120
#> [170,]  58   8      0        161        1   140
#> [171,]  58  14      1          6        0   190
#> [172,]  61   9      1          8        0   150
#> [173,]  58   1      0          1        1   100
#> [174,]  58  10      0         10        1   150
#> [175,]  57   4      1          3        0   138
#> [176,]  61   3      0         17        0   143
#> [177,]  56  14      0         45        0   130
#> [178,]  57   3      1          2        0   120
#> [179,]  56  18      1         11        1   165
#> [180,]  59   9      1          0        1    80
#> [181,]  58  11      0        172        1   135
#> [182,]  60  12      1          0        1   114
#> [183,]  55   9      1          7        1   135
#> [184,]  61  13      1         12        1   130
#> [185,]  59  11      1          8        1   190
#> [186,]  57   1      0          1        0   126
#> [187,]  57  15      1         13        1   110
#> [188,]  59   5      1          2        0   182
#> [189,]  58   5      1          1        1   135
#> [190,]  59  10      0        180        0   160
#> [191,]  61   8      0         77        0   120
#> [192,]  61  13      0         13        0   210
#> [193,]  58   8      1          5        0   152
#> [194,]  62   7      1          2        1   180
#> [195,]  61  18      0        170        0   140
#> [196,]  61  28      1          7        0   133
#> [197,]  58   8      1          3        1   150
#> [198,]  57   7      0        169        0   180
#> [199,]  61   7      0          7        1   150
#> [200,]  60   7      0          7        0   147
#> [201,]  61   6      0          6        0   134
#> [202,]  57  12      1          9        1   120
#> [203,]  62   4      1          0        0   160
#> [204,]  60  17      1          8        1   140
#> [205,]  58   3      1          0        1   146
#> [206,]  59  16      1          9        1   133
#> [207,]  63   6      0         28        1   120
#> [208,]  61  13      0         13        0   120
#> [209,]  61   5      0          5        1   110
#> [210,]  61   5      0          5        1   160
#> [211,]  57   2      1          1        0   159
#> [212,]  62   1      1          0        1   172
#> [213,]  63   3      1          1        0   180
#> [214,]  63   1      0        180        1   130
#> [215,]  61   7      0        180        0   135
#> [216,]  63  15      1         10        1   126
#> [217,]  63   4      1          1        0   155
#> [218,]  59   8      0        180        1   140
#> [219,]  61   9      1          9        1   150
#> [220,]  58   2      0        180        0   127
#> [221,]  59   4      0        180        0   196
#> [222,]  60   7      1          5        1   141
#> [223,]  60   7      0          7        0   140
#> [224,]  59   5      1          1        0   148
#> [225,]  60   7      1          1        1    90
#> [226,]  63   1      0          1        0   162
#> [227,]  63   1      0          1        0   130
#> [228,]  61  15      1         13        0   170
#> [229,]  59   4      0          4        0   149
#> [230,]  63  12      1         10        0   200
#> [231,]  61   6      1          1        1   117
#> [232,]  64  12      1         11        0   160
#> [233,]  66   1      1          0        1   120
#> [234,]  64   6      1          0        1   140
#> [235,]  65  36      1         11        0   140
#> [236,]  63   4      1          3        0   162
#> [237,]  61  10      1          2        1   194
#> [238,]  64  32      1          9        1   160
#> [239,]  63  12      1          9        0   114
#> [240,]  63   7      0        180        0   120
#> [241,]  66   5      1          0        1   110
#> [242,]  65   8      1          0        0   168
#> [243,]  65  10      1          8        1   120
#> [244,]  64   0      0          0        1    90
#> [245,]  60   6      0        180        0   130
#> [246,]  64  21      1         10        0   190
#> [247,]  64   9      0        180        0   150
#> [248,]  65   3      0        180        1   190
#> [249,]  63  16      1          7        1   110
#> [250,]  66   6      1          1        1   130
#> [251,]  63  12      0         12        1   150
#> [252,]  62   3      1          1        1   199
#> [253,]  65   6      0          9        0   112
#> [254,]  63   5      1          4        0   170
#> [255,]  63   2      1          1        0   180
#> [256,]  62  13      1         11        0   180
#> [257,]  67  11      0         11        1   100
#> [258,]  64   2      0          2        0   201
#> [259,]  66  18      1          5        0   142
#> [260,]  62   9      0        180        0   145
#> [261,]  61  15      1         10        0   130
#> [262,]  63   9      1          8        1   160
#> [263,]  64  19      1          8        1   160
#> [264,]  65   8      1          0        1   140
#> [265,]  67   6      0        180        1   170
#> [266,]  68   5      1          4        1   150
#> [267,]  64  13      1         12        1   150
#> [268,]  64   6      1          0        1   125
#> [269,]  64  14      1         13        1   150
#> [270,]  65   3      0          3        0   105
#> [271,]  64   0      0          0        1   148
#> [272,]  67   4      1          3        0   130
#> [273,]  66   3      1          0        1   135
#> [274,]  64  10      1          9        1   110
#> [275,]  63   7      1          0        0   162
#> [276,]  67   8      1          1        1   130
#> [277,]  68   5      0          5        1    90
#> [278,]  66  14      0        180        0   130
#> [279,]  64   1      0          1        1   120
#> [280,]  68  18      0        180        1   260
#> [281,]  65  17      1         14        1   100
#> [282,]  63   8      1          1        1   162
#> [283,]  65  18      1          3        0   120
#> [284,]  63   1      1          0        1   155
#> [285,]  68  11      0        180        0   160
#> [286,]  65  15      1         12        1   150
#> [287,]  66  11      1          0        0   100
#> [288,]  65   4      1          2        1   145
#> [289,]  69  12      0         15        1   140
#> [290,]  66  15      1         13        1   160
#> [291,]  63   2      0        180        0   150
#> [292,]  65  11      1          6        0   130
#> [293,]  69  21      1         10        0   180
#> [294,]  69   6      0        180        1   100
#> [295,]  63   8      0        180        1   120
#> [296,]  65   8      1          0        1    90
#> [297,]  67   1      0        180        1   160
#> [298,]  68  10      1         10        1   150
#> [299,]  65   1      1          0        0   133
#> [300,]  67   7      1          4        1   130
#> [301,]  63   2      1          0        0    99
#> [302,]  67   2      0        180        0   184
#> [303,]  65   6      0          6        0    80
#> [304,]  65  10      1          1        1   148
#> [305,]  66  19      1         12        1   150
#> [306,]  67  12      1         12        0   160
#> [307,]  69   6      0         99        1   140
#> [308,]  65   4      1          1        0   130
#> [309,]  64   4      0        179        0   160
#> [310,]  66   4      0        180        1   130
#> [311,]  70  15      1         12        1   132
#> [312,]  64  11      0         11        0   125
#> [313,]  64   4      0        180        1   140
#> [314,]  64   0      1          0        1   118
#> [315,]  67   2      0         18        0   131
#> [316,]  66   4      0        180        0   177
#> [317,]  69   4      1          3        1   150
#> [318,]  65  13      1         12        1   130
#> [319,]  69  17      1         10        0   140
#> [320,]  69   8      0         93        0   140
#> [321,]  64  21      0         21        1   155
#> [322,]  65   6      0        101        1   115
#> [323,]  70   7      1          0        1   190
#> [324,]  68   7      0        150        0   210
#> [325,]  66   1      1          1        1   165
#> [326,]  69   8      0        180        1   153
#> [327,]  70  14      0        171        0   166
#> [328,]  66   4      0        180        0   130
#> [329,]  68  18      1         14        1   170
#> [330,]  65   2      0        180        0   130
#> [331,]  68   7      1          0        1   150
#> [332,]  69   3      1          2        0   151
#> [333,]  67  14      1         13        0   130
#> [334,]  65  14      1         13        1   150
#> [335,]  71   7      0          7        0   230
#> [336,]  66   2      0          2        1   228
#> [337,]  69   5      0          5        1   142
#> [338,]  71   3      0        103        0   133
#> [339,]  69   3      0          3        1   130
#> [340,]  70  22      1         13        0   103
#> [341,]  67   1      0         36        1   104
#> [342,]  68   6      0        180        0   145
#> [343,]  69   8      1          5        1   195
#> [344,]  69   6      1          4        1   174
#> [345,]  72   7      0          7        1   110
#> [346,]  66   2      1          1        0   123
#> [347,]  69  19      0        180        0   130
#> [348,]  68  18      0         18        1   100
#> [349,]  67  14      0        172        1   140
#> [350,]  69  11      1          0        1   120
#> [351,]  66   2      0        180        0   130
#> [352,]  67   7      1          4        0   122
#> [353,]  69   4      1          3        0   132
#> [354,]  69   8      1          2        0   121
#> [355,]  67  13      1          9        0   130
#> [356,]  70   3      0        123        0   130
#> [357,]  70   9      0        180        1   142
#> [358,]  72   5      1          4        0   170
#> [359,]  67  22      1          1        1   140
#> [360,]  68   3      0         19        0   135
#> [361,]  67  12      1          8        0   120
#> [362,]  69   1      0          1        1   110
#> [363,]  67   1      0          1        1    60
#> [364,]  67   4      0         60        1   136
#> [365,]  69   5      0         76        0   120
#> [366,]  67   8      1          0        1   130
#> [367,]  72  13      1         11        1   195
#> [368,]  68  10      1          8        1   160
#> [369,]  66  24      1         13        0   130
#> [370,]  70  35      1          0        1   105
#> [371,]  72  30      1          0        1   145
#> [372,]  70   7      0          7        0   102
#> [373,]  73  20      1          0        1   170
#> [374,]  71   6      0          9        0   120
#> [375,]  69  10      1          6        1   120
#> [376,]  72  19      1          8        0   120
#> [377,]  72  12      1         10        0   170
#> [378,]  67   8      0        180        1   170
#> [379,]  67   9      0        180        0   158
#> [380,]  73  13      0        152        1   130
#> [381,]  70   5      0        180        0   150
#> [382,]  72   2      0          2        1   100
#> [383,]  67   4      1          1        0   134
#> [384,]  72   6      1          5        0   115
#> [385,]  71   1      0        173        1   188
#> [386,]  68  23      0        180        1   220
#> [387,]  70   3      0        180        0   121
#> [388,]  69   3      0        180        0   220
#> [389,]  71   3      1          2        0   150
#> [390,]  72   5      0         28        0   120
#> [391,]  71   5      0        180        0   191
#> [392,]  69  16      1         10        1   140
#> [393,]  69   8      1          1        0   164
#> [394,]  69   1      1          0        0   155
#> [395,]  73   6      1          0        1   270
#> [396,]  72   8      1          1        1   150
#> [397,]  71   2      1          0        1   180
#> [398,]  73   7      0          7        1   140
#> [399,]  68  15      1         13        1   130
#> [400,]  70   3      0          3        1   159
#> [401,]  70  13      1          9        0   100
#> [402,]  72   6      0        180        1   130
#> [403,]  73   0      0        180        1   161
#> [404,]  73   4      0        180        1   154
#> [405,]  69   2      1          0        1   110
#> [406,]  71   3      1          1        0   150
#> [407,]  71  20      1         10        0   140
#> [408,]  74   0      1          0        1    90
#> [409,]  73   3      1          0        1   136
#> [410,]  71   8      1          7        0   149
#> [411,]  71   3      1          2        1   190
#> [412,]  73  10      1          8        0   106
#> [413,]  69  12      1          1        1   149
#> [414,]  72   5      1          3        1   160
#> [415,]  72  15      1          0        1   150
#> [416,]  71   7      1          2        0   143
#> [417,]  74   3      0          3        1   150
#> [418,]  71  13      1          8        0   121
#> [419,]  70   4      1          0        1   140
#> [420,]  71  14      1         13        1   170
#> [421,]  74   7      1          0        1   117
#> [422,]  72  10      1          8        1   153
#> [423,]  69   7      0        180        1   144
#> [424,]  70   8      0          8        0   120
#> [425,]  71  10      1          9        1   120
#> [426,]  73  10      1         10        1   120
#> [427,]  74  15      1          9        1   179
#> [428,]  71   2      0         10        1   112
#> [429,]  73   1      0          1        1    80
#> [430,]  75   9      1          7        0   140
#> [431,]  75  13      1          1        1   130
#> [432,]  71  11      1          8        0   110
#> [433,]  71   4      0          4        0   134
#> [434,]  72  15      1         12        1   120
#> [435,]  73  10      1          8        0   120
#> [436,]  70   7      1          4        0   184
#> [437,]  72   7      0         57        1   145
#> [438,]  73  10      0        180        0   162
#> [439,]  72  11      0         11        1   140
#> [440,]  70   3      0          3        0   150
#> [441,]  73   5      1          3        1   112
#> [442,]  76  25      1         12        1   170
#> [443,]  73  12      1         12        1   140
#> [444,]  75   1      0        180        1   140
#> [445,]  72   4      1          0        1   197
#> [446,]  71   3      1          0        0   144
#> [447,]  73   5      0        180        0   126
#> [448,]  73   4      0        180        0   124
#> [449,]  74  34      1          8        1   233
#> [450,]  71  32      1         12        1   107
#> [451,]  72   3      0        180        0   160
#> [452,]  77  11      0         11        1   150
#> [453,]  72   7      1          2        0   142
#> [454,]  73  15      0         15        1   160
#> [455,]  71  16      0        180        0   140
#> [456,]  73  10      1         10        0   124
#> [457,]  74   7      0        180        1   150
#> [458,]  74   3      0          3        1   128
#> [459,]  76   1      0        180        0   114
#> [460,]  75  23      1         14        1   110
#> [461,]  74   2      0        180        0   190
#> [462,]  72   4      1          3        0   160
#> [463,]  73   4      1          3        1   125
#> [464,]  75   4      1          0        1   122
#> [465,]  75   7      0          7        0   190
#> [466,]  75   0      0          0        1   130
#> [467,]  75  12      0         12        1   160
#> [468,]  74   6      0        180        0   160
#> [469,]  76   4      0          4        1   155
#> [470,]  74   2      0        180        0   111
#> [471,]  73   1      0         52        1   105
#> [472,]  72   5      0        180        0   120
#> [473,]  78  12      1         11        1   160
#> [474,]  76  44      1         10        0   105
#> [475,]  76   5      1          0        1   167
#> [476,]  74   8      1          8        1   170
#> [477,]  75   9      0        180        1   140
#> [478,]  77   5      1          0        0   123
#> [479,]  77  12      1          9        1   100
#> [480,]  77  12      0        180        0   130
#> [481,]  77   1      1          0        1    90
#> [482,]  76  12      1         11        1   120
#> [483,]  78   5      1          0        1   170
#> [484,]  73   7      1          0        0   174
#> [485,]  74   9      1          8        0   200
#> [486,]  77   3      0        180        0   110
#> [487,]  76  29      0         47        0    90
#> [488,]  74   2      0        180        0   100
#> [489,]  78   7      0          7        1   133
#> [490,]  74  15      0        180        1   172
#> [491,]  74   7      0          7        0   161
#> [492,]  76  13      1          1        1   170
#> [493,]  80  10      1          6        1   147
#> [494,]  78   0      0        180        1   212
#> [495,]  75   5      0        119        1   150
#> [496,]  75  12      1          1        1   120
#> [497,]  75  13      1          6        0   150
#> [498,]  76   1      0          1        1    83
#> [499,]  78  12      1          9        0   150
#> [500,]  78   2      1          1        0   130
#> [501,]  75   4      1          0        0   212
#> [502,]  77   2      1          0        1   143
#> [503,]  75  11      1          4        0   162
#> [504,]  76   7      0         29        1   150
#> [505,]  77  24      0         24        1   160
#> [506,]  79   8      0         32        1   120
#> [507,]  80   9      0         23        1   128
#> [508,]  78   6      1          0        1   240
#> [509,]  78  11      1          1        1   140
#> [510,]  79  11      0        180        0   160
#> [511,]  78  14      1          0        1   140
#> [512,]  78  11      1          8        1   118
#> [513,]  76   4      0          4        1   160
#> [514,]  79   4      0          4        1   125
#> [515,]  76  10      1          8        0   180
#> [516,]  80   3      1          0        1   120
#> [517,]  75   2      1          1        1   204
#> [518,]  78  11      0        180        1   135
#> [519,]  76   1      0          1        1   140
#> [520,]  77  31      1          3        1   161
#> [521,]  76   1      0          1        1    90
#> [522,]  79   3      0          3        0   120
#> [523,]  79   4      1          0        1   120
#> [524,]  81   1      0        180        0   120
#> [525,]  80  15      1         12        1   150
#> [526,]  77   9      1          4        0   141
#> [527,]  82   5      0          8        1   120
#> [528,]  80  17      1         12        0   100
#> [529,]  76   7      0        161        0   151
#> [530,]  81   4      1          2        1   126
#> [531,]  79  28      0        164        0   100
#> [532,]  80   9      0        118        1   186
#> [533,]  80   6      0        173        1   160
#> [534,]  79   1      0         37        1   140
#> [535,]  77  13      1          0        1   190
#> [536,]  78  15      0         15        0   165
#> [537,]  80   5      1          1        1   108
#> [538,]  78  26      1          5        0   194
#> [539,]  76   1      0        166        0   131
#> [540,]  78  20      1          0        1   109
#> [541,]  80   1      0          1        0   100
#> [542,]  78   3      1          1        1   152
#> [543,]  77  10      1          8        1   130
#> [544,]  82   3      1          1        1   144
#> [545,]  77   5      0         85        0   188
#> [546,]  79   6      0          6        0   152
#> [547,]  80   6      1          0        1   119
#> [548,]  80   5      0          5        1   130
#> [549,]  82   1      1          0        1    82
#> [550,]  79  10      0        180        1   150
#> [551,]  77   4      0        180        1    98
#> [552,]  78  12      0        180        0   134
#> [553,]  79   1      0        125        0   193
#> [554,]  84  22      1         10        0   180
#> [555,]  79   4      0          4        1   121
#> [556,]  80   6      0          6        1   110
#> [557,]  83   9      1          5        1   170
#> [558,]  82   5      0        180        0   110
#> [559,]  79   7      1          6        0   130
#> [560,]  83   4      0        103        0    97
#> [561,]  81   5      0        177        0    41
#> [562,]  80  11      1          8        0   170
#> [563,]  78  23      1         10        1   145
#> [564,]  82   8      1          1        0   128
#> [565,]  79   1      0        180        1   170
#> [566,]  81  15      0        180        1   140
#> [567,]  80   7      1          0        1   146
#> [568,]  81  20      1          9        0   170
#> [569,]  80   6      1          0        1   150
#> [570,]  80  11      1          8        0   110
#> [571,]  81   8      0        180        0   146
#> [572,]  80   8      1          7        0   160
#> [573,]  81   2      1          1        0   198
#> [574,]  83   2      0          2        1   155
#> [575,]  82   6      0        128        1   100
#> [576,]  84   4      0        167        0   198
#> [577,]  82  23      1          0        0   110
#> [578,]  81   1      0          1        1   150
#> [579,]  84   1      0         38        1   205
#> [580,]  83   3      0        180        0   174
#> [581,]  81   4      0         90        1   138
#> [582,]  79   9      1          8        0   150
#> [583,]  85   3      1          2        1   160
#> [584,]  80  13      1          8        1   140
#> [585,]  80   2      1          0        1   130
#> [586,]  79   4      0          4        1    60
#> [587,]  82  19      0         19        0   120
#> [588,]  80  30      1         13        0   220
#> [589,]  83   9      0        180        0   198
#> [590,]  79  14      1          0        0   110
#> [591,]  81  14      1         12        1   128
#> [592,]  83   2      0        154        0   130
#> [593,]  82   0      0          2        1   100
#> [594,]  85   9      1          6        1   160
#> [595,]  83   1      0        180        0   160
#> [596,]  84  15      1         13        1   110
#> [597,]  81   1      0          1        1   145
#> [598,]  81   4      0          4        0   160
#> [599,]  83  12      1          2        1   170
#> [600,]  81  19      1         14        0   120
#> [601,]  82   3      1          2        0   130
#> [602,]  83   7      0        126        0   135
#> [603,]  84   6      0        165        0   145
#> [604,]  82   9      0        180        1   134
#> [605,]  81  13      0        180        0   152
#> [606,]  85   3      0          3        1   118
#> [607,]  81   4      0        180        0   160
#> [608,]  83   9      0        180        1   149
#> [609,]  82   1      0        180        1   193
#> [610,]  83   4      0          4        0   130
#> [611,]  87   2      0          5        1   137
#> [612,]  86  12      1          0        1   132
#> [613,]  82  14      1         11        1   103
#> [614,]  84  16      0         70        1   150
#> [615,]  83  19      0         43        0   150
#> [616,]  86   2      0        180        1   169
#> [617,]  88  14      1          3        1   130
#> [618,]  84   7      1          2        0   148
#> [619,]  84   9      0         92        1   110
#> [620,]  84   3      0        180        1   170
#> [621,]  86   4      0         38        1   122
#> [622,]  82   4      0          4        0   130
#> [623,]  86  13      0        177        0   163
#> [624,]  85   3      0          3        1   113
#> [625,]  86   6      0          6        1   117
#> [626,]  84  13      0         62        1   100
#> [627,]  86   6      1          1        0   112
#> [628,]  88   4      0          4        1   115
#> [629,]  86   6      0         46        0   173
#> [630,]  88   3      0        115        0   110
#> [631,]  87   8      0          8        1   157
#> [632,]  88   4      0          4        0    86
#> [633,]  87   6      0        180        1   110
#> [634,]  84   8      0        180        1   119
#> [635,]  87  29      0         29        1    97
#> [636,]  87  15      1          9        1   138
#> [637,]  84   0      0        180        1   136
#> [638,]  89  10      0         46        1   170
#> [639,]  90  14      0         14        1   100
#> [640,]  88   1      0          1        0   135
#> [641,]  91  10      0        145        0   135
#> [642,]  86   3      1          0        1    80
#> [643,]  88   8      0         50        1   154
#> [644,]  90  11      1         10        1   186
#> [645,]  87   6      0        126        1   168
#> [646,]  86   9      1          7        0   130
#> [647,]  90   4      1          0        0   121
#> [648,]  87  43      0        178        1   130
#> [649,]  90   5      1          0        1   125
#> [650,]  88   3      1          2        0   159
#> [651,]  89   3      1          1        1   160
#> [652,]  92   1      0          1        1   167
#> [653,]  88   5      0        158        0   100
#> [654,]  87   7      0         74        1   105
#> [655,]  91   5      0        169        1   176
#> [656,]  89  52      0         52        1   130
#> [657,]  89   4      0          4        1   159
#> [658,]  91   0      0          0        0     0
#> [659,]  91   4      1          0        1   120
#> [660,]  90   1      0          1        1   118
#> [661,]  91   2      0          2        1   116
#> [662,]  94   8      0          8        1   142
#> [663,]  92   4      0         76        1   149
#> [664,]  90   3      0         67        0   162
#> [665,]  96   3      0         12        1    97
#> [666,]  95   8      1          5        1   150
#> [667,]  94   3      0         26        1   144
#> [668,]  91   7      0          7        0   135
#> [669,]  92   5      0         69        0   139
#> [670,]  96  15      1          0        1   140
#> 
#> $y
#>   [1]   5.0+ 180.0+   5.0+ 180.0+ 180.0+ 180.0+  12.0    5.0+ 180.0+ 180.0+
#>  [11] 180.0+ 180.0+ 180.0+   2.0+   2.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [21] 180.0+ 180.0+ 180.0+ 180.0+ 155.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0+
#>  [31] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 150.0  180.0+ 180.0+   6.0+ 180.0+
#>  [41] 180.0+ 180.0+  73.0    5.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0+ 180.0+
#>  [51] 180.0+ 180.0+ 177.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [61] 180.0+ 180.0+ 180.0+ 180.0+   2.0    1.0  179.0+ 180.0+ 180.0+ 180.0+
#>  [71] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   7.0   36.0   88.0+ 180.0+ 180.0+
#>  [81] 180.0+ 180.0+  77.0  180.0+ 180.0+ 180.0+   5.0  180.0+ 180.0+ 180.0+
#>  [91] 180.0+   4.0+ 180.0+ 166.0+  16.0+ 152.0+   7.0+ 180.0+ 180.0+  13.0+
#> [101] 180.0+ 174.0+  28.0    6.0+   1.0  180.0+   9.0+ 180.0+ 180.0+ 180.0+
#> [111] 175.0+   2.0  180.0+   7.0+ 180.0+ 180.0+ 180.0+ 180.0+  16.0+ 180.0+
#> [121]  15.0+ 180.0+ 180.0+ 180.0+   8.0    2.0    3.0+ 180.0+ 180.0+ 180.0+
#> [131]   1.0  165.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [141] 180.0+ 180.0+ 171.0+  15.0    5.0+ 147.0+ 180.0+ 180.0+   5.0+ 180.0+
#> [151]   4.0+ 180.0+   9.0+   1.0  180.0+ 180.0+   2.0+ 180.0+ 180.0+ 180.0+
#> [161]   2.0  180.0+ 180.0+ 180.0+ 180.0+  64.0  180.0+   9.0+ 180.0+ 161.0+
#> [171] 171.0+ 180.0+   1.0   10.0+ 180.0+  17.0   45.0    3.0+ 180.0+   9.0+
#> [181] 172.0+ 172.0+  24.0  180.0+ 180.0+   1.0+  15.0  180.0+ 180.0+ 180.0+
#> [191]  77.0   13.0+   8.0+ 180.0+ 170.0   94.0  180.0+ 169.0    7.0    7.0+
#> [201]   6.0  180.0+ 180.0+ 180.0+   3.0+ 180.0+  28.0   13.0+   5.0    5.0+
#> [211] 180.0+   1.0  180.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+ 180.0+ 180.0+
#> [221] 180.0+  84.0    7.0+ 180.0+ 180.0+   1.0    1.0  180.0+   4.0+ 180.0+
#> [231] 180.0+  12.0  180.0+ 180.0+  36.0  180.0+  88.0  180.0+  12.0  180.0+
#> [241] 180.0+ 180.0+ 180.0+   0.5  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [251]  12.0  180.0+   9.0  180.0+ 180.0+ 180.0+  11.0+   2.0+  18.0+ 180.0+
#> [261] 180.0+ 180.0+ 103.0   15.0  180.0+   5.0+  13.0  180.0+  14.0+   3.0 
#> [271]   0.5+ 180.0+   3.0+ 180.0+   7.0+   8.0    5.0  180.0+   1.0  180.0+
#> [281] 180.0+ 180.0+ 123.0+   1.0+ 180.0+  15.0+ 180.0+   4.0+  15.0  180.0+
#> [291] 180.0+ 180.0+ 174.0+ 180.0+ 180.0+   8.0+ 180.0+  10.0  180.0+ 180.0+
#> [301] 180.0+ 180.0+   6.0  180.0+  19.0+  12.0   99.0  180.0+ 179.0+ 180.0+
#> [311] 180.0+  11.0+ 180.0+   0.5   18.0  180.0+ 152.0+ 180.0+ 180.0+  93.0 
#> [321]  21.0+ 101.0    7.0+ 150.0    1.0  180.0+ 171.0  180.0+ 180.0+ 180.0+
#> [331] 180.0+ 180.0+ 180.0+  14.0+   7.0+   2.0    5.0+ 103.0    3.0+ 180.0+
#> [341]  36.0  180.0+ 180.0+  97.0    7.0    2.0+ 180.0+  18.0  172.0+ 180.0+
#> [351] 180.0+   7.0  180.0+   8.0+  13.0+ 123.0  180.0+ 180.0+  51.0   19.0 
#> [361] 180.0+   1.0    1.0   60.0   76.0  180.0+ 132.0   10.0+ 180.0+ 180.0+
#> [371] 162.0    7.0+ 124.0    9.0  180.0+ 180.0+  12.0  180.0+ 180.0+ 152.0 
#> [381] 180.0+   2.0   76.0  180.0+ 173.0+ 180.0+ 180.0+ 180.0+ 180.0+  28.0 
#> [391] 180.0+  16.0+ 180.0+ 180.0+   6.0  180.0+ 180.0+   7.0+  15.0    3.0+
#> [401]  13.0+ 180.0+ 180.0+ 180.0+   2.0    3.0+  20.0    0.5  180.0+   8.0 
#> [411]   3.0   87.0   12.0  180.0+ 180.0+ 180.0+   3.0  175.0  180.0+  14.0+
#> [421] 180.0+  10.0+ 180.0+   8.0+ 179.0+  15.0  180.0+  10.0    1.0  180.0+
#> [431]  13.0  180.0+   4.0+ 180.0+  10.0  104.0+  57.0  180.0+  11.0    3.0+
#> [441]   5.0  180.0+  12.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+  34.0  177.0+
#> [451] 180.0+  11.0+   7.0   15.0+ 180.0+  10.0  180.0+   3.0  180.0+ 180.0+
#> [461] 180.0+ 180.0+ 180.0+   4.0    7.0    0.5   12.0  180.0+   4.0  180.0+
#> [471]  52.0  180.0+  12.0  180.0+ 180.0+   8.0  180.0+   5.0  180.0+ 180.0+
#> [481]   1.0   12.0  180.0+   7.0+ 168.0+ 180.0+  47.0  180.0+   7.0  180.0+
#> [491]   7.0  180.0+  10.0  180.0+ 119.0   12.0  180.0+   1.0  180.0+ 180.0+
#> [501]   4.0+   2.0  152.0+  29.0   24.0   32.0   23.0  180.0+ 180.0+ 180.0+
#> [511] 180.0+  11.0    4.0    4.0   10.0+   3.0+   2.0+ 180.0+   1.0  171.0 
#> [521]   1.0    3.0  138.0  180.0+ 180.0+  71.0    8.0   17.0  161.0   93.0 
#> [531] 164.0  118.0  173.0   37.0   22.0   15.0+   5.0+ 171.0+ 166.0+  20.0+
#> [541]   1.0    3.0+  10.0  180.0+  85.0    6.0+   6.0    5.0    1.0  180.0+
#> [551] 180.0+ 180.0+ 125.0  180.0+   4.0    6.0    9.0+ 180.0+ 180.0+ 103.0 
#> [561] 177.0+ 169.0   70.0  180.0+ 180.0+ 180.0+   7.0+  20.0  180.0+ 180.0+
#> [571] 180.0+ 180.0+ 180.0+   2.0  128.0  167.0   62.0    1.0   38.0  180.0+
#> [581]  90.0  180.0+ 180.0+ 180.0+ 180.0+   4.0   19.0   30.0  180.0+ 180.0+
#> [591] 180.0+ 154.0    2.0  180.0+ 180.0+ 180.0+   1.0    4.0+  77.0  180.0+
#> [601]   3.0  126.0  165.0  180.0+ 180.0+   3.0+ 180.0+ 180.0+ 180.0+   4.0+
#> [611]   5.0  180.0+ 174.0   70.0   43.0  180.0+  14.0  180.0+  92.0  180.0+
#> [621]  38.0    4.0  177.0    3.0+   6.0+  62.0    6.0+   4.0   46.0  115.0 
#> [631]   8.0+   4.0  180.0+ 180.0+  29.0  180.0+ 180.0+  46.0   14.0    1.0+
#> [641] 145.0    3.0   50.0   11.0  126.0  180.0+   4.0  178.0+  89.0   75.0 
#> [651]   3.0+   1.0  158.0   74.0  169.0   52.0    4.0    0.5    4.0    1.0+
#> [661]   2.0    8.0+  76.0   67.0   12.0    8.0   26.0    7.0+  69.0   15.0+
#> 
#> $weights
#> NULL
#> 
#> $offset
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8281484