Cross-Validated GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::cv.glmnet() from package glmnet.

Details

This learner returns two prediction types:

lp: a vector of linear predictors (relative risk scores), one per observation. Calculated using glmnet::predict.cv.glmnet().

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.cv.glmnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used.

Custom mlr3 parameters

family is set to "cox" and cannot be changed.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.cv_glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

foldid	untyped	NULL		-

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nfolds	integer	10		\([3, \infty)\)

nlambda	integer	100		\([1, \infty)\)

offset	untyped	NULL		-

newoffset	untyped	-		-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	lambda.1se		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

standardize.response	logical	FALSE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.gaussian	character	-	covariance, naive	-

type.logistic	character	Newton	Newton, modified.Newton	-

type.measure	character	deviance	deviance, C	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvCVGlmnet

Methods

Public methods

LearnerSurvCVGlmnet$new()

LearnerSurvCVGlmnet$selected_features()

LearnerSurvCVGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvCVGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvCVGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvCVGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.cv_glmnet")
print(learner)
#> <LearnerSurvCVGlmnet:surv.cv_glmnet>: Regularized Generalized Linear Model
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3proba, mlr3extralearners, glmnet
#> * Predict Types:  [crank], distr, lp
#> * Feature Types: logical, integer, numeric
#> * Properties: selected_features, weights

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#> Measure: Partial Likelihood Deviance 
#> 
#>      Lambda Index Measure     SE Nonzero
#> min 0.00335    44   8.742 0.2482       6
#> 1se 0.04982    15   8.967 0.2538       5
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  28   9      0        180        1   107
#>   [2,]  32   5      1          0        1   121
#>   [3,]  33   2      0          2        0   150
#>   [4,]  35   5      1          2        0   172
#>   [5,]  35  10      1          9        0   106
#>   [6,]  34   5      0          5        0   120
#>   [7,]  35   2      0        180        0   121
#>   [8,]  38   2      0        115        0   150
#>   [9,]  36   1      0        180        1   155
#>  [10,]  38  12      1          8        1   120
#>  [11,]  33   6      1          1        1   115
#>  [12,]  38  16      1         10        0   160
#>  [13,]  40  12      1          9        0   153
#>  [14,]  42   3      1          1        1   130
#>  [15,]  37   1      1          0        1   146
#>  [16,]  40   2      1          1        1   148
#>  [17,]  42   2      0        180        1   100
#>  [18,]  42   2      0          2        0   140
#>  [19,]  40  11      1         10        1   120
#>  [20,]  42   2      0        180        0   100
#>  [21,]  40   1      1          0        1   145
#>  [22,]  43   4      1          0        1   130
#>  [23,]  42   4      0        180        0   162
#>  [24,]  40   3      1          1        0   170
#>  [25,]  42  12      1         10        1   170
#>  [26,]  43   2      1          1        1   116
#>  [27,]  42   2      0        180        1   124
#>  [28,]  44   5      1          1        0   170
#>  [29,]  45   3      0        180        1   154
#>  [30,]  45   9      1          7        0   110
#>  [31,]  45   6      0        180        1   170
#>  [32,]  41   5      1          4        1   141
#>  [33,]  43   2      0        180        1   140
#>  [34,]  46  15      0        180        0   120
#>  [35,]  48  15      0        180        1   160
#>  [36,]  45   3      0        150        0   130
#>  [37,]  46   7      1          2        0   166
#>  [38,]  43  29      0        180        1   180
#>  [39,]  45   4      1          0        0   124
#>  [40,]  43  10      0        180        0   185
#>  [41,]  47   6      1          0        1   116
#>  [42,]  46  13      1         10        0   100
#>  [43,]  47   4      1          3        1   160
#>  [44,]  43   3      1          0        1   124
#>  [45,]  45   8      1          0        1   117
#>  [46,]  49   5      0         73        1   136
#>  [47,]  45   5      0          5        0   141
#>  [48,]  46   2      1          1        1   122
#>  [49,]  46   6      1          0        1   100
#>  [50,]  44   4      1          0        1   114
#>  [51,]  44   9      1          8        1   135
#>  [52,]  45   5      0        180        1   190
#>  [53,]  46   5      1          3        0   130
#>  [54,]  46   4      0        180        1   121
#>  [55,]  46  15      0        180        1   120
#>  [56,]  48  12      1         11        0   200
#>  [57,]  47   9      1          6        0   170
#>  [58,]  46   3      1          0        1   119
#>  [59,]  47  10      0         10        1   140
#>  [60,]  48   2      1          0        0   184
#>  [61,]  50   4      1          1        0   125
#>  [62,]  49   7      1          7        1   110
#>  [63,]  46   3      1          1        1   140
#>  [64,]  46   9      1          9        1   122
#>  [65,]  50   7      0        180        1   110
#>  [66,]  49   2      0          2        0   105
#>  [67,]  51   1      0          1        1   145
#>  [68,]  47   2      0        180        0   150
#>  [69,]  49  23      0        179        1   112
#>  [70,]  50   7      1          0        1    92
#>  [71,]  50   4      0          4        1   100
#>  [72,]  51   3      1          2        0   113
#>  [73,]  50   1      1          0        0   150
#>  [74,]  50   9      0        180        0   130
#>  [75,]  49   7      1          4        1    90
#>  [76,]  47   8      0        180        0   160
#>  [77,]  47   6      0        180        1   162
#>  [78,]  46   3      0        180        1   120
#>  [79,]  46   1      1          1        0   145
#>  [80,]  48  17      1         10        0   111
#>  [81,]  52   4      1          4        0   152
#>  [82,]  49   9      1          3        0   102
#>  [83,]  49  15      0        180        1   160
#>  [84,]  53   5      0        180        1   140
#>  [85,]  54  17      1         12        1   102
#>  [86,]  51   3      1          1        0   140
#>  [87,]  50  14      1         13        0   170
#>  [88,]  53   8      1          7        0   160
#>  [89,]  48   3      1          2        0   150
#>  [90,]  51  25      1          1        0   202
#>  [91,]  49   5      1          2        1   150
#>  [92,]  52  14      1          7        1   200
#>  [93,]  53   4      1          0        1   156
#>  [94,]  49  16      0         16        0   125
#>  [95,]  54  23      1         10        0   131
#>  [96,]  55   6      1          2        1   114
#>  [97,]  54   9      1          1        0   130
#>  [98,]  55   4      1          2        0   150
#>  [99,]  52   4      0        180        1   180
#> [100,]  51  13      1         11        0   145
#> [101,]  50   5      1          4        1   150
#> [102,]  55  28      1         13        1   160
#> [103,]  49   6      1          0        1   130
#> [104,]  49   1      0          1        1   110
#> [105,]  53   8      1          0        1   130
#> [106,]  50   7      1          0        1   127
#> [107,]  56   4      1          1        1   130
#> [108,]  55   1      0        180        0   127
#> [109,]  55   2      0          2        0   145
#> [110,]  54   1      0        180        0   162
#> [111,]  56   3      0        180        1   193
#> [112,]  53  18      1          9        1   150
#> [113,]  54   3      0        180        1   180
#> [114,]  55   6      0        180        0   170
#> [115,]  52  16      0         16        0   152
#> [116,]  53  10      1          9        0   172
#> [117,]  52  16      1         14        0   170
#> [118,]  55   6      1          5        1   138
#> [119,]  56   3      0          8        1   139
#> [120,]  55   1      0          2        0   130
#> [121,]  54   7      1          2        0   129
#> [122,]  54   2      1          1        0   135
#> [123,]  52   9      1          3        0   170
#> [124,]  57   5      1          3        1   138
#> [125,]  57   1      0          1        1   100
#> [126,]  56   4      1          0        1   140
#> [127,]  55  11      1          7        0   104
#> [128,]  53   3      1          0        1   200
#> [129,]  57  10      0        180        1   170
#> [130,]  54   5      0        180        1   108
#> [131,]  55   3      1          1        1   156
#> [132,]  53  21      1         13        1   130
#> [133,]  59   3      1          1        0   172
#> [134,]  57   4      0        180        1   119
#> [135,]  58   6      1          0        1    90
#> [136,]  53  15      1         10        1   130
#> [137,]  54  17      1          8        1   227
#> [138,]  55   9      1          2        1   147
#> [139,]  55  13      0        166        1   140
#> [140,]  56   5      0          5        1   150
#> [141,]  54  23      1          8        0   120
#> [142,]  57   4      1          2        1   185
#> [143,]  53   7      1          0        1   120
#> [144,]  57  11      1         10        1   129
#> [145,]  55   3      1          2        0   140
#> [146,]  55   5      0          5        1   131
#> [147,]  54   7      1          0        1   141
#> [148,]  56   4      0          4        0   164
#> [149,]  58   9      1          0        1   180
#> [150,]  55   5      1          0        0   140
#> [151,]  56   7      1          5        1   120
#> [152,]  55   2      0          2        0   106
#> [153,]  59   9      1          1        1   125
#> [154,]  60  11      1          9        0   106
#> [155,]  59   3      0        180        0   120
#> [156,]  58   4      1          0        1   160
#> [157,]  57   2      0          2        1   120
#> [158,]  60   5      1          1        0   138
#> [159,]  55   5      1          0        1   160
#> [160,]  57  10      1          9        0   103
#> [161,]  59   6      1          0        1   140
#> [162,]  59   4      1          0        1   152
#> [163,]  58  26      1          0        1   189
#> [164,]  58   4      1          3        0   120
#> [165,]  60   0      1          0        1    80
#> [166,]  59   2      1          1        0   140
#> [167,]  58   8      0        161        1   140
#> [168,]  58  14      1          6        0   190
#> [169,]  61   4      1          3        0   151
#> [170,]  61   9      1          8        0   150
#> [171,]  61   3      1          2        1   102
#> [172,]  58   1      0          1        1   100
#> [173,]  57   4      1          3        0   138
#> [174,]  56  14      0         45        0   130
#> [175,]  56  18      1         11        1   165
#> [176,]  55   4      1          3        1   160
#> [177,]  58  11      0        172        1   135
#> [178,]  56   8      1          8        0   120
#> [179,]  61  13      1         12        1   130
#> [180,]  59  11      1          8        1   190
#> [181,]  57   1      0          1        0   126
#> [182,]  57  15      1         13        1   110
#> [183,]  61   8      0         77        0   120
#> [184,]  62   7      1          2        1   180
#> [185,]  57   3      1          0        0   100
#> [186,]  61  18      0        170        0   140
#> [187,]  58   8      1          3        1   150
#> [188,]  57   7      0        169        0   180
#> [189,]  61   7      0          7        1   150
#> [190,]  60   7      0          7        0   147
#> [191,]  59  13      1          2        0   198
#> [192,]  57  12      1          9        1   120
#> [193,]  62   4      1          0        0   160
#> [194,]  58   3      1          0        1   146
#> [195,]  62   4      1          3        0   173
#> [196,]  59  16      1          9        1   133
#> [197,]  61  13      0         13        0   120
#> [198,]  61   5      0          5        1   160
#> [199,]  58  11      1          9        0   179
#> [200,]  62   1      1          0        1   172
#> [201,]  58   7      0        180        1   150
#> [202,]  63   3      1          1        0   180
#> [203,]  61   7      0        180        0   135
#> [204,]  63   4      1          3        0   222
#> [205,]  63   4      0        180        1   190
#> [206,]  63  15      1         10        1   126
#> [207,]  64   4      0        180        0   130
#> [208,]  63   4      1          1        0   155
#> [209,]  60  18      1         13        0   132
#> [210,]  59   8      0        180        1   140
#> [211,]  61   9      1          9        1   150
#> [212,]  58   9      1          9        0   110
#> [213,]  62   7      0          7        0   150
#> [214,]  59   1      0         22        1   162
#> [215,]  58   2      0        180        0   127
#> [216,]  59   4      0        180        0   196
#> [217,]  60   7      1          5        1   141
#> [218,]  60   7      0          7        0   140
#> [219,]  65  13      0        180        1   100
#> [220,]  63   1      0          1        0   162
#> [221,]  63   1      0          1        0   130
#> [222,]  62   6      0        180        0   170
#> [223,]  61  15      1         13        0   170
#> [224,]  59   4      0          4        0   149
#> [225,]  60   3      0          3        0   168
#> [226,]  64  10      1          9        0   160
#> [227,]  62   6      0          6        0   120
#> [228,]  59  10      0        180        1   130
#> [229,]  61   6      1          1        1   117
#> [230,]  64  12      1         11        0   160
#> [231,]  65  36      1         11        0   140
#> [232,]  63   4      1          3        0   162
#> [233,]  63  12      1          9        0   114
#> [234,]  63   7      0        180        0   120
#> [235,]  66   5      1          0        1   110
#> [236,]  65   8      1          0        0   168
#> [237,]  65  10      1          8        1   120
#> [238,]  64   0      0          0        1    90
#> [239,]  60   6      0        180        0   130
#> [240,]  64  21      1         10        0   190
#> [241,]  61  12      1         11        0   154
#> [242,]  64   9      0        180        0   150
#> [243,]  61   4      0        180        1   113
#> [244,]  65   3      0        180        1   190
#> [245,]  63  16      1          7        1   110
#> [246,]  63  12      0         12        1   150
#> [247,]  65   6      0          9        0   112
#> [248,]  65   3      1          0        1    80
#> [249,]  63   5      1          4        0   170
#> [250,]  63   2      1          1        0   180
#> [251,]  62  13      1         11        0   180
#> [252,]  67  11      0         11        1   100
#> [253,]  66  16      1         11        1   169
#> [254,]  61  14      1          5        0   140
#> [255,]  63   9      1          8        1   160
#> [256,]  63   3      1          2        0   120
#> [257,]  64  19      1          8        1   160
#> [258,]  65   8      1          0        1   140
#> [259,]  68   5      1          4        1   150
#> [260,]  64  13      1         12        1   150
#> [261,]  66   7      1          0        1   115
#> [262,]  66  13      1          0        0   118
#> [263,]  64  14      1         13        1   150
#> [264,]  65   3      0          3        0   105
#> [265,]  64   0      0          0        1   148
#> [266,]  67   4      1          3        0   130
#> [267,]  66   6      1          0        1   140
#> [268,]  64  10      1          9        1   110
#> [269,]  63   7      1          0        0   162
#> [270,]  68   5      0          5        1    90
#> [271,]  63  10      0         16        1   160
#> [272,]  66  14      0        180        0   130
#> [273,]  64   1      0          1        1   120
#> [274,]  68  18      0        180        1   260
#> [275,]  65  17      1         14        1   100
#> [276,]  63   8      1          1        1   162
#> [277,]  65  18      1          3        0   120
#> [278,]  63   1      1          0        1   155
#> [279,]  63  10      0         18        1   130
#> [280,]  67  11      0         11        0   150
#> [281,]  68  11      0        180        0   160
#> [282,]  68  14      0         79        0   172
#> [283,]  66  12      1         10        1   150
#> [284,]  65  15      1         12        1   150
#> [285,]  66  11      1          0        0   100
#> [286,]  65   4      1          2        1   145
#> [287,]  69  12      0         15        1   140
#> [288,]  66  15      1         13        1   160
#> [289,]  63   2      0        180        0   150
#> [290,]  65  11      1          6        0   130
#> [291,]  69  21      1         10        0   180
#> [292,]  66   9      1          8        0   130
#> [293,]  68  14      1         13        1   140
#> [294,]  69   1      1          0        0   170
#> [295,]  68  10      1         10        1   150
#> [296,]  67   7      1          4        1   130
#> [297,]  63   2      1          0        0    99
#> [298,]  67   2      0        180        0   184
#> [299,]  65   6      0          6        0    80
#> [300,]  66  19      1         12        1   150
#> [301,]  67  12      1         12        0   160
#> [302,]  66   4      0        180        1   130
#> [303,]  70  15      1         12        1   132
#> [304,]  64  11      0         11        0   125
#> [305,]  64   4      0        180        1   140
#> [306,]  64   0      1          0        1   118
#> [307,]  67   2      0         18        0   131
#> [308,]  66   7      1          5        1   131
#> [309,]  66   4      0        180        0   177
#> [310,]  69   4      1          3        1   150
#> [311,]  65  13      1         12        1   130
#> [312,]  64  21      0         21        1   155
#> [313,]  66   6      0        180        0   140
#> [314,]  65   1      0          1        1   120
#> [315,]  65   6      0        101        1   115
#> [316,]  68   4      0          4        1   190
#> [317,]  71   3      0          5        0   112
#> [318,]  70   7      1          0        1   190
#> [319,]  71  20      1          0        1   160
#> [320,]  67   2      0        180        0   128
#> [321,]  66   9      1          3        1   151
#> [322,]  66   1      1          1        1   165
#> [323,]  70   4      1          0        1   180
#> [324,]  69   8      0        180        1   153
#> [325,]  67   6      1          4        0   130
#> [326,]  68  18      1         14        1   170
#> [327,]  69   0      0          0        1   148
#> [328,]  65   2      0        180        0   130
#> [329,]  68   7      1          0        1   150
#> [330,]  69   3      1          2        0   151
#> [331,]  67  14      1         13        0   130
#> [332,]  65  14      1         13        1   150
#> [333,]  69   8      0        180        1   180
#> [334,]  71   7      0          7        0   230
#> [335,]  66   2      0          2        1   228
#> [336,]  69   5      0          5        1   142
#> [337,]  71   3      0        103        0   133
#> [338,]  70  22      1         13        0   103
#> [339,]  67   1      0         36        1   104
#> [340,]  67   5      0          5        0   130
#> [341,]  68   6      0        180        0   145
#> [342,]  69   8      1          5        1   195
#> [343,]  72   3      1          0        1   132
#> [344,]  72   7      0          7        1   110
#> [345,]  69   8      1          7        1   108
#> [346,]  67   3      0        180        0   110
#> [347,]  66   2      1          1        0   123
#> [348,]  69  19      0        180        0   130
#> [349,]  67  14      0        172        1   140
#> [350,]  69  11      1          0        1   120
#> [351,]  66   2      0        180        0   130
#> [352,]  67   7      1          4        0   122
#> [353,]  68   2      0          7        1   130
#> [354,]  70   9      0        180        1   142
#> [355,]  72   5      1          4        0   170
#> [356,]  68   3      0         19        0   135
#> [357,]  67  12      1          8        0   120
#> [358,]  69   1      0          1        1   110
#> [359,]  67   1      0          1        1    60
#> [360,]  67   4      0         60        1   136
#> [361,]  67   8      1          0        1   130
#> [362,]  72  13      1         11        1   195
#> [363,]  68  10      1          8        1   160
#> [364,]  66  24      1         13        0   130
#> [365,]  70  35      1          0        1   105
#> [366,]  72  30      1          0        1   145
#> [367,]  68   7      1          2        0   135
#> [368,]  73  20      1          0        1   170
#> [369,]  71   6      0          9        0   120
#> [370,]  69  10      1          6        1   120
#> [371,]  70  11      0        180        1   210
#> [372,]  72  19      1          8        0   120
#> [373,]  73  13      0        152        1   130
#> [374,]  70   5      0        180        0   150
#> [375,]  72   2      0          2        1   100
#> [376,]  71   1      0        173        1   188
#> [377,]  68  23      0        180        1   220
#> [378,]  70   3      0        180        0   121
#> [379,]  69   3      0        180        0   220
#> [380,]  71   3      1          2        0   150
#> [381,]  72   5      0         28        0   120
#> [382,]  71   5      0        180        0   191
#> [383,]  69  16      1         10        1   140
#> [384,]  69   8      1          1        0   164
#> [385,]  72  16      1          1        1   130
#> [386,]  70   4      0        180        0   180
#> [387,]  69   1      1          0        0   155
#> [388,]  73   7      0          7        1   140
#> [389,]  68  15      1         13        1   130
#> [390,]  70  13      1          9        0   100
#> [391,]  72   6      0        180        1   130
#> [392,]  73   0      0        180        1   161
#> [393,]  74   8      1          0        1    85
#> [394,]  73   4      0        180        1   154
#> [395,]  71   3      1          1        0   150
#> [396,]  71  15      1         11        0   165
#> [397,]  68   9      0        180        1   120
#> [398,]  71  20      1         10        0   140
#> [399,]  73   3      1          0        1   136
#> [400,]  70   5      1          0        1   190
#> [401,]  71   8      1          7        0   149
#> [402,]  71   3      1          2        1   190
#> [403,]  73  10      1          8        0   106
#> [404,]  72   5      1          3        1   160
#> [405,]  70   3      0        180        1   154
#> [406,]  73   6      0        180        0   110
#> [407,]  72  15      1          0        1   150
#> [408,]  72   8      1          0        1   140
#> [409,]  74   3      0          3        1   150
#> [410,]  71  13      1          8        0   121
#> [411,]  69   2      1          1        1    80
#> [412,]  71  14      1         13        1   170
#> [413,]  74   7      1          0        1   117
#> [414,]  69   7      0        180        1   144
#> [415,]  72  15      1         13        0   156
#> [416,]  75   1      0          1        0   133
#> [417,]  75   2      1          1        0   145
#> [418,]  74  15      1          9        1   179
#> [419,]  71   2      0         10        1   112
#> [420,]  73   1      0          1        1    80
#> [421,]  75   9      1          7        0   140
#> [422,]  71  11      1          8        0   110
#> [423,]  71   4      0          4        0   134
#> [424,]  72   1      1          1        0   168
#> [425,]  73  10      0        180        0   162
#> [426,]  70   3      0          3        0   150
#> [427,]  73   5      1          3        1   112
#> [428,]  73  12      1         12        1   140
#> [429,]  75   1      0        180        1   140
#> [430,]  72   4      1          0        1   197
#> [431,]  71   3      1          0        0   144
#> [432,]  73   5      0        180        0   126
#> [433,]  73   4      0        180        0   124
#> [434,]  74  34      1          8        1   233
#> [435,]  72   3      0        180        0   160
#> [436,]  77  11      0         11        1   150
#> [437,]  77   4      0          4        0   185
#> [438,]  72   7      1          2        0   142
#> [439,]  73  15      0         15        1   160
#> [440,]  74   7      0        180        1   150
#> [441,]  74   3      0          3        1   128
#> [442,]  76   1      0        180        0   114
#> [443,]  74   2      1          1        0   140
#> [444,]  76   8      1          0        1   141
#> [445,]  74   2      0        180        0   190
#> [446,]  72   4      0         85        1   120
#> [447,]  72   4      1          3        0   160
#> [448,]  76  17      1          0        1   200
#> [449,]  73   4      1          3        1   125
#> [450,]  76  13      1         10        0   110
#> [451,]  75   4      1          0        1   122
#> [452,]  75   0      0          0        1   130
#> [453,]  73  13      1         11        0   195
#> [454,]  75  12      0         12        1   160
#> [455,]  74   8      1          0        1   105
#> [456,]  75   4      1          2        1   188
#> [457,]  74   6      0        180        0   160
#> [458,]  76   4      0          4        1   155
#> [459,]  75   1      0          1        1   125
#> [460,]  74   2      0        180        0   111
#> [461,]  73   0      0        180        0   156
#> [462,]  72   5      0        180        0   120
#> [463,]  78  12      1         11        1   160
#> [464,]  74  10      1          0        1   135
#> [465,]  76   5      1          0        1   167
#> [466,]  74   8      1          8        1   170
#> [467,]  75   9      0        180        1   140
#> [468,]  73  33      1         12        1   175
#> [469,]  77   5      1          0        0   123
#> [470,]  77  12      1          9        1   100
#> [471,]  77  12      0        180        0   130
#> [472,]  78   5      1          0        1   170
#> [473,]  73   7      1          0        0   174
#> [474,]  74   6      0         79        1   140
#> [475,]  75   3      1          1        1   171
#> [476,]  79  10      1          8        0   190
#> [477,]  74   2      1          0        1   130
#> [478,]  78  18      0         18        1   144
#> [479,]  76  29      0         47        0    90
#> [480,]  74   2      0        180        0   100
#> [481,]  78   7      0          7        1   133
#> [482,]  74  15      0        180        1   172
#> [483,]  78   8      1          6        1   110
#> [484,]  78  32      1          9        1   198
#> [485,]  79   6      0        180        0   170
#> [486,]  80  10      1          6        1   147
#> [487,]  78   0      0        180        1   212
#> [488,]  78  13      1          5        0   130
#> [489,]  75  12      1          1        1   120
#> [490,]  78  15      0        180        1   270
#> [491,]  80   8      0          8        1   120
#> [492,]  75  13      1          6        0   150
#> [493,]  74  10      1          8        0   135
#> [494,]  78  12      1          9        0   150
#> [495,]  78   2      1          1        0   130
#> [496,]  75   4      1          0        0   212
#> [497,]  77   2      1          0        1   143
#> [498,]  78  10      0        180        1   130
#> [499,]  76   7      0         29        1   150
#> [500,]  77  24      0         24        1   160
#> [501,]  80   9      0         23        1   128
#> [502,]  80   6      0          6        1   150
#> [503,]  78   6      1          0        1   240
#> [504,]  76   3      1          0        1   140
#> [505,]  78  11      1          1        1   140
#> [506,]  79  11      0        180        0   160
#> [507,]  79   2      1          0        1   121
#> [508,]  78  14      1          0        1   140
#> [509,]  76   4      0          4        1   160
#> [510,]  76  12      1         10        1   127
#> [511,]  75   2      1          1        1   204
#> [512,]  78   7      1          0        1   110
#> [513,]  77   7      0        180        1   170
#> [514,]  77   6      0          6        1   144
#> [515,]  79   4      1          0        1   120
#> [516,]  81   1      0        180        0   120
#> [517,]  82   5      0          8        1   120
#> [518,]  80  40      1          0        1   138
#> [519,]  80  17      1         12        0   100
#> [520,]  76   7      0        161        0   151
#> [521,]  79  10      0         10        1   120
#> [522,]  81   4      1          2        1   126
#> [523,]  79  28      0        164        0   100
#> [524,]  80   6      0        173        1   160
#> [525,]  78  32      0        180        1   130
#> [526,]  81   3      0        180        0   184
#> [527,]  81   2      0        175        0   172
#> [528,]  78   7      0          7        1   147
#> [529,]  80   5      1          1        1   108
#> [530,]  79   3      0          3        1   101
#> [531,]  76   1      0        166        0   131
#> [532,]  81   4      1          1        1   104
#> [533,]  78   3      1          1        1   152
#> [534,]  77  10      1          8        1   130
#> [535,]  82   3      1          1        1   144
#> [536,]  77   5      0         85        0   188
#> [537,]  80   2      1          1        0   168
#> [538,]  79   6      0          6        0   152
#> [539,]  80   6      1          0        1   119
#> [540,]  78   2      0        180        0   148
#> [541,]  80   5      0          5        1   130
#> [542,]  82   1      1          0        1    82
#> [543,]  79  10      0        180        1   150
#> [544,]  77   4      0        180        1    98
#> [545,]  81   1      0        108        0   129
#> [546,]  82  21      1          2        0   155
#> [547,]  84  22      1         10        0   180
#> [548,]  79   4      0          4        1   121
#> [549,]  80   6      0          6        1   110
#> [550,]  83   9      1          5        1   170
#> [551,]  82   5      0        180        0   110
#> [552,]  83   5      0        180        0   148
#> [553,]  83   4      0        103        0    97
#> [554,]  81   5      0        177        0    41
#> [555,]  78  23      1         10        1   145
#> [556,]  78   9      1          4        1   120
#> [557,]  82   8      1          1        0   128
#> [558,]  79   1      0        180        1   170
#> [559,]  80   7      1          0        1   146
#> [560,]  84   5      1          1        1    85
#> [561,]  83   8      0          8        0   115
#> [562,]  80   6      1          0        1   150
#> [563,]  80  11      1          8        0   110
#> [564,]  81   8      0        180        0   146
#> [565,]  80   8      1          7        0   160
#> [566,]  79   0      1          0        1    96
#> [567,]  83   2      0          2        1   155
#> [568,]  82   6      0        128        1   100
#> [569,]  84   4      0        167        0   198
#> [570,]  82  23      1          0        0   110
#> [571,]  84   1      0         38        1   205
#> [572,]  83   3      0        180        0   174
#> [573,]  79   9      1          8        0   150
#> [574,]  85   3      1          2        1   160
#> [575,]  84   4      0         89        1   129
#> [576,]  80   6      0         71        1   189
#> [577,]  80  30      1         13        0   220
#> [578,]  79  14      1          0        0   110
#> [579,]  83   2      0        154        0   130
#> [580,]  85   9      1          6        1   160
#> [581,]  83   1      0        180        0   160
#> [582,]  84  15      1         13        1   110
#> [583,]  81  12      0         12        1   163
#> [584,]  83  12      1          2        1   170
#> [585,]  81  19      1         14        0   120
#> [586,]  82   3      1          2        0   130
#> [587,]  82  15      1          0        0   183
#> [588,]  83   7      0        126        0   135
#> [589,]  86   8      0          8        1   132
#> [590,]  84   6      0        165        0   145
#> [591,]  86   3      0          3        1   140
#> [592,]  82   9      0        180        1   134
#> [593,]  84   3      0        180        1   120
#> [594,]  85   3      0          3        1   118
#> [595,]  81   2      1          0        1   118
#> [596,]  83   9      0        180        1   149
#> [597,]  82   1      0        180        1   193
#> [598,]  83   4      0          4        0   130
#> [599,]  87   2      0          5        1   137
#> [600,]  86  12      1          0        1   132
#> [601,]  86   6      1          0        1   140
#> [602,]  84  16      0         70        1   150
#> [603,]  83  19      0         43        0   150
#> [604,]  84   3      1          2        0   125
#> [605,]  83  10      1          0        1   190
#> [606,]  84   7      1          2        0   148
#> [607,]  84   9      0         92        1   110
#> [608,]  84   3      0        180        1   170
#> [609,]  86   4      0         38        1   122
#> [610,]  85   3      0          3        1   113
#> [611,]  84  13      0         62        1   100
#> [612,]  86   6      1          1        0   112
#> [613,]  88   4      0          4        0   100
#> [614,]  83  20      1          3        1   150
#> [615,]  88   4      0          4        1   115
#> [616,]  85  22      0         22        1   184
#> [617,]  83   9      0         65        1   150
#> [618,]  87   2      0        180        1   130
#> [619,]  86   6      0         46        0   173
#> [620,]  88   3      0        115        0   110
#> [621,]  83   3      0          3        1   130
#> [622,]  87   8      0          8        1   157
#> [623,]  86  15      1          8        1   109
#> [624,]  88   4      0          4        0    86
#> [625,]  89   4      0          4        1   153
#> [626,]  89   5      0        119        1   140
#> [627,]  87   6      0        180        1   110
#> [628,]  87   1      0          1        0   170
#> [629,]  84   2      0        110        1   174
#> [630,]  87  29      0         29        1    97
#> [631,]  89  10      0         46        1   170
#> [632,]  90  14      0         14        1   100
#> [633,]  88   1      0          1        0   135
#> [634,]  86   4      0        180        1   145
#> [635,]  87   6      1          0        0   125
#> [636,]  91  10      0        145        0   135
#> [637,]  86   3      1          0        1    80
#> [638,]  90  11      1         10        1   186
#> [639,]  87   6      0        126        1   168
#> [640,]  86  10      0        180        1   137
#> [641,]  86   9      1          7        0   130
#> [642,]  91   1      0          1        1    74
#> [643,]  87  43      0        178        1   130
#> [644,]  87   5      0         36        1   150
#> [645,]  88   3      1          2        0   159
#> [646,]  92   1      0          1        1   167
#> [647,]  91   3      0         33        1   137
#> [648,]  88   5      0        158        0   100
#> [649,]  87   7      0         74        1   105
#> [650,]  89  12      1          0        1   130
#> [651,]  91   5      0        169        1   176
#> [652,]  89  52      0         52        1   130
#> [653,]  92   7      0          7        1   110
#> [654,]  89   4      0          4        1   159
#> [655,]  89  14      0        180        1    84
#> [656,]  90  19      1         11        1   129
#> [657,]  94   6      0         50        0    78
#> [658,]  91   2      0          2        1   116
#> [659,]  93   8      0        179        1   110
#> [660,]  92   4      0         76        1   149
#> [661,]  90   3      0         67        0   162
#> [662,]  95   8      1          5        1   150
#> [663,]  94   3      0         26        1   144
#> [664,]  91  12      0         53        1   212
#> [665,]  93   0      1          0        1   122
#> [666,]  92   5      0         69        0   139
#> [667,]  92   2      0          2        0   112
#> [668,]  93   4      0        180        1   135
#> [669,]  96   3      1          0        1   104
#> [670,]  96  15      1          0        1   140
#> 
#> $y
#>   [1] 180.0+   5.0+   2.0+   5.0+ 180.0+   5.0+ 180.0+ 115.0  180.0+  12.0 
#>  [11] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 180.0+   2.0+ 180.0+ 180.0+
#>  [21] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 180.0+ 155.0+ 180.0+ 180.0+
#>  [31] 180.0+   5.0+ 180.0+ 180.0+ 180.0+ 150.0  180.0+ 180.0+ 180.0+ 180.0+
#>  [41]   6.0+ 180.0+ 180.0+ 180.0+ 180.0+  73.0    5.0+ 161.0+ 180.0+ 180.0+
#>  [51] 180.0+ 180.0+   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  10.0+ 180.0+
#>  [61] 180.0+   7.0  180.0+ 180.0+ 180.0+   2.0    1.0  180.0+ 179.0+ 180.0+
#>  [71]   4.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  88.0+
#>  [81]   4.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [91] 180.0+  85.0  166.0+  16.0+ 152.0+   6.0+ 180.0+ 180.0+ 180.0+  13.0+
#> [101] 171.0+  28.0    6.0+   1.0  180.0+ 180.0+ 180.0+ 180.0+   2.0  180.0+
#> [111] 180.0+ 180.0+ 180.0+ 180.0+  16.0+ 180.0+  16.0  180.0+   8.0    2.0 
#> [121] 180.0+ 180.0+ 180.0+ 140.0    1.0  165.0  180.0+ 180.0+ 180.0+ 180.0+
#> [131] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 171.0+  15.0  166.0+   5.0+
#> [141] 180.0+   4.0+ 180.0+ 180.0+ 180.0+   5.0+ 180.0+   4.0+   9.0+ 180.0+
#> [151] 180.0+   2.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0  180.0+ 180.0+ 180.0+
#> [161]  64.0  180.0+ 180.0+ 180.0+   0.5  180.0+ 161.0+ 171.0+ 180.0+ 180.0+
#> [171]   3.0    1.0  180.0+  45.0  180.0+ 180.0+ 172.0+   8.0  180.0+ 180.0+
#> [181]   1.0+  15.0   77.0  180.0+ 180.0+ 170.0  180.0+ 169.0    7.0    7.0+
#> [191] 180.0+ 180.0+ 180.0+   3.0+ 180.0+ 180.0+  13.0+   5.0+ 180.0+   1.0 
#> [201] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+ 180.0+
#> [211] 180.0+   9.0    7.0+  22.0  180.0+ 180.0+  84.0    7.0+ 180.0+   1.0 
#> [221]   1.0  180.0+ 180.0+   4.0+   3.0+ 167.0    6.0+ 180.0+ 180.0+  12.0 
#> [231]  36.0  180.0+  12.0  180.0+ 180.0+ 180.0+ 180.0+   0.5  180.0+ 180.0+
#> [241]  12.0+ 180.0+ 180.0+ 180.0+ 180.0+  12.0    9.0    3.0  180.0+ 180.0+
#> [251] 180.0+  11.0+ 180.0+ 180.0+ 180.0+   3.0+ 103.0   15.0    5.0+  13.0 
#> [261] 179.0+ 166.0+  14.0+   3.0    0.5+ 180.0+ 180.0+ 180.0+   7.0+   5.0 
#> [271]  16.0  180.0+   1.0  180.0+ 180.0+ 180.0+ 123.0+   1.0+  18.0   11.0+
#> [281] 180.0+  79.0   80.0   15.0+ 180.0+   4.0+  15.0  180.0+ 180.0+ 180.0+
#> [291] 174.0+ 180.0+ 180.0+ 175.0   10.0  180.0+ 180.0+ 180.0+   6.0   19.0+
#> [301]  12.0  180.0+ 180.0+  11.0+ 180.0+   0.5   18.0    7.0+ 180.0+ 152.0+
#> [311] 180.0+  21.0+ 180.0+   1.0  101.0    4.0    5.0    7.0+ 180.0+ 180.0+
#> [321] 180.0+   1.0  180.0+ 180.0+   6.0  180.0+   0.5  180.0+ 180.0+ 180.0+
#> [331] 180.0+  14.0+ 180.0+   7.0+   2.0    5.0+ 103.0  180.0+  36.0    5.0+
#> [341] 180.0+ 180.0+ 180.0+   7.0    8.0+ 180.0+   2.0+ 180.0+ 172.0+ 180.0+
#> [351] 180.0+   7.0    7.0  180.0+ 180.0+  19.0  180.0+   1.0    1.0   60.0 
#> [361] 180.0+ 132.0   10.0+ 180.0+ 180.0+ 162.0    7.0+ 124.0    9.0  180.0+
#> [371] 180.0+ 180.0+ 152.0  180.0+   2.0  173.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [381]  28.0  180.0+  16.0+ 180.0+  16.0+ 180.0+ 180.0+   7.0+  15.0   13.0+
#> [391] 180.0+ 180.0+ 180.0+ 180.0+   3.0+ 180.0+ 180.0+  20.0  180.0+ 180.0+
#> [401]   8.0    3.0   87.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+   3.0  175.0 
#> [411]   2.0   14.0+ 180.0+ 180.0+ 180.0+   1.0  180.0+ 180.0+  10.0    1.0 
#> [421] 180.0+ 180.0+   4.0+   1.0  180.0+   3.0+   5.0   12.0  180.0+ 180.0+
#> [431] 180.0+ 180.0+ 180.0+  34.0  180.0+  11.0+   4.0+   7.0   15.0+ 180.0+
#> [441]   3.0  180.0+ 180.0+ 180.0+ 180.0+  85.0  180.0+  17.0+ 180.0+ 174.0+
#> [451]   4.0    0.5  180.0+  12.0  180.0+  46.0  180.0+   4.0    1.0  180.0+
#> [461] 180.0+ 180.0+  12.0  180.0+ 180.0+   8.0  180.0+  33.0    5.0  180.0+
#> [471] 180.0+ 180.0+   7.0+  79.0    3.0  180.0+ 176.0+  18.0   47.0  180.0+
#> [481]   7.0  180.0+   8.0+  32.0  180.0+  10.0  180.0+ 172.0   12.0  180.0+
#> [491]   8.0  180.0+ 180.0+ 180.0+ 180.0+   4.0+   2.0  180.0+  29.0   24.0 
#> [501]  23.0    6.0  180.0+   3.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0  180.0+
#> [511]   2.0+  43.0  180.0+   6.0  138.0  180.0+   8.0   40.0   17.0  161.0 
#> [521]  10.0+  93.0  164.0  173.0  180.0+ 180.0+ 175.0+   7.0+   5.0+   3.0 
#> [531] 166.0+  71.0    3.0+  10.0  180.0+  85.0   10.0    6.0+   6.0  180.0+
#> [541]   5.0    1.0  180.0+ 180.0+ 108.0  180.0+ 180.0+   4.0    6.0    9.0+
#> [551] 180.0+ 180.0+ 103.0  177.0+  70.0  180.0+ 180.0+ 180.0+   7.0+ 180.0+
#> [561]   8.0+ 180.0+ 180.0+ 180.0+ 180.0+   0.5    2.0  128.0  167.0   62.0 
#> [571]  38.0  180.0+ 180.0+ 180.0+  89.0   71.0   30.0  180.0+ 154.0  180.0+
#> [581] 180.0+ 180.0+  12.0   77.0  180.0+   3.0   83.0  126.0    8.0  165.0 
#> [591]   3.0  180.0+ 180.0+   3.0+ 180.0+ 180.0+ 180.0+   4.0+   5.0  180.0+
#> [601]   6.0   70.0   43.0  180.0+ 180.0+ 180.0+  92.0  180.0+  38.0    3.0+
#> [611]  62.0    6.0+   4.0+  20.0    4.0   22.0   65.0  180.0+  46.0  115.0 
#> [621]   3.0+   8.0+ 180.0+   4.0    4.0  119.0  180.0+   1.0+ 110.0   29.0 
#> [631]  46.0   14.0    1.0+ 180.0+  25.0  145.0    3.0   11.0  126.0  180.0+
#> [641] 180.0+   1.0  178.0+  36.0   75.0    1.0   33.0  158.0   74.0  180.0+
#> [651] 169.0   52.0    7.0    4.0  180.0+ 180.0+  50.0    2.0  179.0+  76.0 
#> [661]  67.0    8.0   26.0   53.0    0.5   69.0    2.0  180.0+   3.0   15.0+
#> 
#> $weights
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> surv.cindex 
#>    0.861766