Cross-Validated GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::cv.glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.cv.glmnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.cv.glmnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.cv_glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

foldid	untyped	NULL		-

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nfolds	integer	10		\([3, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	TRUE	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	lambda.1se		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

standardize.response	logical	FALSE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.gaussian	character	-	covariance, naive	-

type.logistic	character	Newton	Newton, modified.Newton	-

type.measure	character	deviance	deviance, C	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.coxnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvCVGlmnet

Methods

Public methods

LearnerSurvCVGlmnet$new()

LearnerSurvCVGlmnet$selected_features()

LearnerSurvCVGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvCVGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvCVGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvCVGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.cv_glmnet")
print(learner)
#> 
#> ── <LearnerSurvCVGlmnet> (surv.cv_glmnet): Regularized Generalized Linear Model 
#> • Model: -
#> • Parameters: use_pred_offset=TRUE
#> • Packages: mlr3, mlr3proba, mlr3extralearners, and glmnet
#> • Predict Types: [crank], distr, and lp
#> • Feature Types: logical, integer, and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: offset, selected_features, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#> Measure: Partial Likelihood Deviance 
#> 
#>      Lambda Index Measure     SE Nonzero
#> min 0.00515    40   2.945 0.2127       6
#> 1se 0.09211     9   3.148 0.2136       3
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  28   9      0        180        1   107
#>   [2,]  32   5      1          0        1   121
#>   [3,]  33   2      0          2        0   150
#>   [4,]  34   5      0          5        0   120
#>   [5,]  35   2      0        180        0   121
#>   [6,]  35   2      1          1        1   112
#>   [7,]  38  13      1          0        1   161
#>   [8,]  38   2      0        115        0   150
#>   [9,]  36   5      1          0        1   115
#>  [10,]  33   6      1          1        1   115
#>  [11,]  38  16      1         10        0   160
#>  [12,]  38  12      1         11        1    92
#>  [13,]  40  12      1          9        0   153
#>  [14,]  42   3      1          1        1   130
#>  [15,]  37   1      1          0        1   146
#>  [16,]  42   2      0        180        1   100
#>  [17,]  42   2      0          2        0   140
#>  [18,]  40   6      0        180        1   138
#>  [19,]  40  11      1         10        1   120
#>  [20,]  42   2      0        180        0   100
#>  [21,]  43   3      1          0        1   100
#>  [22,]  40   1      1          0        1   145
#>  [23,]  42   4      0        180        0   162
#>  [24,]  42  15      1         13        1   125
#>  [25,]  42  12      1         10        1   170
#>  [26,]  42   2      0        180        1   124
#>  [27,]  41  10      1          8        0   150
#>  [28,]  44   3      0        180        0   141
#>  [29,]  41  13      1          1        0   140
#>  [30,]  41   5      1          4        1   141
#>  [31,]  44   2      1          1        1   150
#>  [32,]  43   2      0        180        1   140
#>  [33,]  45   2      0        180        1   140
#>  [34,]  46  15      0        180        0   120
#>  [35,]  46   2      1          1        0   126
#>  [36,]  47   4      1          3        0   118
#>  [37,]  44   3      1          0        1   180
#>  [38,]  46   7      1          2        0   166
#>  [39,]  43  29      0        180        1   180
#>  [40,]  45   4      1          0        0   124
#>  [41,]  43  10      0        180        0   185
#>  [42,]  47   6      1          0        1   116
#>  [43,]  46  13      1         10        0   100
#>  [44,]  44   0      1          0        1    96
#>  [45,]  47   4      1          3        1   160
#>  [46,]  45   8      1          0        1   117
#>  [47,]  46   6      1          0        1   100
#>  [48,]  47   2      0        180        0   108
#>  [49,]  44   9      1          8        1   135
#>  [50,]  45   5      0        180        1   190
#>  [51,]  46   4      0        180        1   121
#>  [52,]  44   2      0        180        0   142
#>  [53,]  48   3      0        180        0   154
#>  [54,]  48  12      1         11        0   200
#>  [55,]  47   5      1          3        1   130
#>  [56,]  46   3      1          0        1   119
#>  [57,]  49   4      0        180        0   117
#>  [58,]  47  10      0         10        1   140
#>  [59,]  50   1      1          0        1   129
#>  [60,]  49   7      1          7        1   110
#>  [61,]  46   9      1          9        1   122
#>  [62,]  50   7      0        180        1   110
#>  [63,]  49   2      0          2        0   105
#>  [64,]  51   1      0          1        1   145
#>  [65,]  47   2      0        180        0   150
#>  [66,]  49  23      0        179        1   112
#>  [67,]  46   6      1          0        1   156
#>  [68,]  52   2      0        180        1   170
#>  [69,]  50   7      1          0        1    92
#>  [70,]  50   1      1          0        0   150
#>  [71,]  50   9      0        180        0   130
#>  [72,]  51   8      0        180        1   140
#>  [73,]  52   2      0        180        0   155
#>  [74,]  46   1      1          1        0   145
#>  [75,]  50   4      1          1        0   150
#>  [76,]  48   7      1          0        1   110
#>  [77,]  48  17      1         10        0   111
#>  [78,]  47   2      1          1        0   110
#>  [79,]  52   4      1          4        0   152
#>  [80,]  49   9      1          3        0   102
#>  [81,]  53   5      0        180        1   140
#>  [82,]  54  17      1         12        1   102
#>  [83,]  53   7      1          0        0   199
#>  [84,]  51   3      1          1        0   140
#>  [85,]  50  10      1          6        0   122
#>  [86,]  53   8      1          7        0   160
#>  [87,]  48   3      1          2        0   150
#>  [88,]  49   5      1          2        1   150
#>  [89,]  52  14      1          7        1   200
#>  [90,]  48   6      0        180        0   160
#>  [91,]  53   4      1          0        1   156
#>  [92,]  51  13      0         99        1   160
#>  [93,]  54   9      1          0        1   138
#>  [94,]  55   3      1          1        0   150
#>  [95,]  54  23      1         10        0   131
#>  [96,]  54   9      1          1        0   130
#>  [97,]  55   4      1          2        0   150
#>  [98,]  52   4      0        180        1   180
#>  [99,]  50   5      1          4        1   150
#> [100,]  50   3      0        174        1   153
#> [101,]  55  28      1         13        1   160
#> [102,]  49   6      1          0        1   130
#> [103,]  53   9      0          9        1    95
#> [104,]  50   7      1          0        1   127
#> [105,]  52   5      0        175        1   117
#> [106,]  55   1      0        180        0   127
#> [107,]  55   2      0          2        0   145
#> [108,]  54   7      1          0        1   100
#> [109,]  56   3      0        180        1   193
#> [110,]  56   2      0        180        0   132
#> [111,]  55   5      1          4        1   120
#> [112,]  53  18      1          9        1   150
#> [113,]  55   6      0        180        0   170
#> [114,]  52  16      0         16        0   152
#> [115,]  52  16      1         14        0   170
#> [116,]  53  15      0         15        1    90
#> [117,]  53   4      0        180        1   150
#> [118,]  55   6      1          5        1   138
#> [119,]  54  12      1          0        1   190
#> [120,]  55   2      0        134        1   140
#> [121,]  57   3      0          3        0   120
#> [122,]  54   7      1          2        0   129
#> [123,]  54   2      1          1        0   135
#> [124,]  54   2      1          1        1   176
#> [125,]  57   5      1          3        1   138
#> [126,]  57   1      0        180        1   156
#> [127,]  57   1      0          1        1   100
#> [128,]  56   4      1          0        1   140
#> [129,]  52   2      0        180        0   140
#> [130,]  52  15      1         14        0   130
#> [131,]  56  14      1         11        0   130
#> [132,]  53   3      1          0        1   200
#> [133,]  57  10      0        180        1   170
#> [134,]  58   8      0          8        1   130
#> [135,]  54   5      0        180        1   108
#> [136,]  57   0      0          0        1   150
#> [137,]  57   4      0        180        1   119
#> [138,]  53  15      1         10        1   130
#> [139,]  54  17      1          8        1   227
#> [140,]  55   9      1          2        1   147
#> [141,]  55  13      0        166        1   140
#> [142,]  54  23      1          8        0   120
#> [143,]  57   4      1          2        1   185
#> [144,]  53   7      1          0        1   120
#> [145,]  55   3      1          2        0   140
#> [146,]  55   5      0          5        1   131
#> [147,]  54   7      1          0        1   141
#> [148,]  56   4      0          4        0   164
#> [149,]  59  15      1         10        0   140
#> [150,]  58   9      1          0        1   180
#> [151,]  55   5      1          0        0   140
#> [152,]  55   2      0          2        0   106
#> [153,]  59   9      1          1        1   125
#> [154,]  57   1      0        180        0   148
#> [155,]  60  11      1          9        0   106
#> [156,]  59   3      0        180        0   120
#> [157,]  57   2      0          2        1   120
#> [158,]  60   5      1          1        0   138
#> [159,]  57   5      0        180        1   130
#> [160,]  58  11      1          9        1   124
#> [161,]  59   6      1          0        1   140
#> [162,]  59   5      0        180        1   155
#> [163,]  59   4      1          0        1   152
#> [164,]  60   0      1          0        1    80
#> [165,]  59   2      1          1        0   140
#> [166,]  61   4      1          3        0   151
#> [167,]  61   3      1          2        1   102
#> [168,]  58   1      0          1        1   100
#> [169,]  61  20      1         13        0   130
#> [170,]  57   2      1          0        1   116
#> [171,]  58  10      0         10        1   150
#> [172,]  57  11      0        180        1   150
#> [173,]  61   3      0         17        0   143
#> [174,]  56  14      0         45        0   130
#> [175,]  57   3      1          2        0   120
#> [176,]  58  19      1         13        1   140
#> [177,]  56  13      1          6        1   158
#> [178,]  56  18      1         11        1   165
#> [179,]  59   9      1          0        1    80
#> [180,]  55   4      1          3        1   160
#> [181,]  60  12      1          0        1   114
#> [182,]  61   4      1          0        1   115
#> [183,]  56   8      1          8        0   120
#> [184,]  61  13      1         12        1   130
#> [185,]  57  15      1         13        1   110
#> [186,]  59   5      1          2        0   182
#> [187,]  58   5      1          1        1   135
#> [188,]  59  10      0        180        0   160
#> [189,]  61   8      0         77        0   120
#> [190,]  58   8      1          5        0   152
#> [191,]  62  10      1          0        1   153
#> [192,]  62   7      1          2        1   180
#> [193,]  61  28      1          7        0   133
#> [194,]  58   8      1          3        1   150
#> [195,]  57   7      0        169        0   180
#> [196,]  61   7      0          7        1   150
#> [197,]  60   7      0          7        0   147
#> [198,]  61   6      0          6        0   134
#> [199,]  57  12      1          9        1   120
#> [200,]  58   3      1          0        1   146
#> [201,]  62   4      1          3        0   173
#> [202,]  63   6      0         28        1   120
#> [203,]  61   5      0          5        1   110
#> [204,]  57  18      1          9        1    93
#> [205,]  58  11      1          9        0   179
#> [206,]  57   2      1          1        0   159
#> [207,]  62  17      1         10        1   180
#> [208,]  62   1      1          0        1   172
#> [209,]  58   7      0        180        1   150
#> [210,]  63   3      1          1        0   180
#> [211,]  63   1      0        180        1   130
#> [212,]  61   7      0        180        0   135
#> [213,]  63   4      1          3        0   222
#> [214,]  62   3      0        180        1   105
#> [215,]  63  15      1         10        1   126
#> [216,]  60  18      1         13        0   132
#> [217,]  59   8      0        180        1   140
#> [218,]  61   9      1          9        1   150
#> [219,]  58   9      1          9        0   110
#> [220,]  62   7      0          7        0   150
#> [221,]  59   1      0         22        1   162
#> [222,]  59   4      0        180        0   196
#> [223,]  60   7      1          5        1   141
#> [224,]  60   7      1          1        1    90
#> [225,]  65  13      0        180        1   100
#> [226,]  63   1      0          1        0   130
#> [227,]  61  15      1         13        0   170
#> [228,]  60   3      0          3        0   168
#> [229,]  64  10      1          9        0   160
#> [230,]  63  12      1         10        0   200
#> [231,]  59  10      0        180        1   130
#> [232,]  64  12      1         11        0   160
#> [233,]  64   6      1          0        1   140
#> [234,]  63  10      1          0        1   148
#> [235,]  65  36      1         11        0   140
#> [236,]  66   3      1          1        0   127
#> [237,]  61  10      1          2        1   194
#> [238,]  64  32      1          9        1   160
#> [239,]  63  12      1          9        0   114
#> [240,]  63   7      0        180        0   120
#> [241,]  66   5      1          0        1   110
#> [242,]  64   0      0          0        1    90
#> [243,]  60   6      0        180        0   130
#> [244,]  61  12      1         11        0   154
#> [245,]  64   9      0        180        0   150
#> [246,]  64   7      0        180        1   120
#> [247,]  66   6      1          1        1   130
#> [248,]  63  12      0         12        1   150
#> [249,]  65   6      0          9        0   112
#> [250,]  62  13      1         11        0   180
#> [251,]  67  11      0         11        1   100
#> [252,]  64   2      0          2        0   201
#> [253,]  66  18      1          5        0   142
#> [254,]  66  16      1         11        1   169
#> [255,]  62   9      0        180        0   145
#> [256,]  61  15      1         10        0   130
#> [257,]  63   9      1          8        1   160
#> [258,]  63   3      1          2        0   120
#> [259,]  63   2      1          0        0   140
#> [260,]  64  19      1          8        1   160
#> [261,]  65   8      1          0        1   140
#> [262,]  68   5      1          4        1   150
#> [263,]  64   6      1          0        1   125
#> [264,]  66  13      1          0        0   118
#> [265,]  64  14      1         13        1   150
#> [266,]  64   0      0          0        1   148
#> [267,]  67   4      1          3        0   130
#> [268,]  66   3      1          0        1   135
#> [269,]  66   6      1          0        1   140
#> [270,]  65   2      1          1        1   170
#> [271,]  68   1      0        180        1   166
#> [272,]  67   8      1          1        1   130
#> [273,]  68   5      0          5        1    90
#> [274,]  63  10      0         16        1   160
#> [275,]  64   1      0          1        1   120
#> [276,]  65  17      1         14        1   100
#> [277,]  65  18      1          3        0   120
#> [278,]  63   1      1          0        1   155
#> [279,]  68  11      0        180        0   160
#> [280,]  68  14      0         79        0   172
#> [281,]  66  11      1          0        0   100
#> [282,]  65   4      1          2        1   145
#> [283,]  69  12      0         15        1   140
#> [284,]  66  15      1         13        1   160
#> [285,]  63   2      0        180        0   150
#> [286,]  65  11      1          6        0   130
#> [287,]  69  21      1         10        0   180
#> [288,]  69   6      0        180        1   100
#> [289,]  65   8      1          0        1    90
#> [290,]  69   1      1          0        0   170
#> [291,]  68  10      1         10        1   150
#> [292,]  65   1      1          0        0   133
#> [293,]  63   2      1          0        0    99
#> [294,]  67   2      0        180        0   184
#> [295,]  65   6      0          6        0    80
#> [296,]  65  10      1          1        1   148
#> [297,]  69   6      0         99        1   140
#> [298,]  64   4      0        179        0   160
#> [299,]  70  15      1         12        1   132
#> [300,]  64   4      0        180        1   140
#> [301,]  64   0      1          0        1   118
#> [302,]  67   2      0         18        0   131
#> [303,]  69   4      1          3        1   150
#> [304,]  69   8      0         93        0   140
#> [305,]  66   6      0        180        0   140
#> [306,]  65   1      0          1        1   120
#> [307,]  68  18      1          0        1   160
#> [308,]  65   6      0        101        1   115
#> [309,]  68   4      0          4        1   190
#> [310,]  71   3      0          5        0   112
#> [311,]  71  20      1          0        1   160
#> [312,]  67   2      0        180        0   128
#> [313,]  70   4      1          0        1   180
#> [314,]  66   4      0        180        0   130
#> [315,]  67  10      1          9        0   200
#> [316,]  65   2      0        180        0   130
#> [317,]  68   7      1          0        1   150
#> [318,]  69   3      1          2        0   151
#> [319,]  67  14      1         13        0   130
#> [320,]  71   7      0          7        0   230
#> [321,]  66   2      0          2        1   228
#> [322,]  71   6      0         45        1   158
#> [323,]  69   5      0          5        1   142
#> [324,]  71   3      0        103        0   133
#> [325,]  70  22      1         13        0   103
#> [326,]  67   1      0         36        1   104
#> [327,]  67   5      0          5        0   130
#> [328,]  72   3      1          0        1   132
#> [329,]  72   7      0          7        1   110
#> [330,]  69   8      1          7        1   108
#> [331,]  67   3      0        180        0   110
#> [332,]  66   2      1          1        0   123
#> [333,]  68  18      0         18        1   100
#> [334,]  67  14      0        172        1   140
#> [335,]  69  11      1          0        1   120
#> [336,]  66   2      0        180        0   130
#> [337,]  67   7      1          4        0   122
#> [338,]  69   4      1          3        0   132
#> [339,]  68   2      0          7        1   130
#> [340,]  69   8      1          2        0   121
#> [341,]  67  13      1          9        0   130
#> [342,]  70   9      0        180        1   142
#> [343,]  72   5      1          4        0   170
#> [344,]  67  22      1          1        1   140
#> [345,]  68   3      0         19        0   135
#> [346,]  67  12      1          8        0   120
#> [347,]  69   1      0          1        1   110
#> [348,]  72  13      1         11        1   195
#> [349,]  66  24      1         13        0   130
#> [350,]  73  20      1          0        1   170
#> [351,]  71   6      0          9        0   120
#> [352,]  70  11      0        180        1   210
#> [353,]  72  19      1          8        0   120
#> [354,]  72  12      1         10        0   170
#> [355,]  67   9      0        180        0   158
#> [356,]  73  13      0        152        1   130
#> [357,]  70   5      0        180        0   150
#> [358,]  72   2      0          2        1   100
#> [359,]  67   4      1          1        0   134
#> [360,]  71   1      0        173        1   188
#> [361,]  70   3      0        180        0   121
#> [362,]  68   4      1          3        0   210
#> [363,]  72   5      0         28        0   120
#> [364,]  71   5      0        180        0   191
#> [365,]  68   7      0        180        1   130
#> [366,]  72  16      1          1        1   130
#> [367,]  69   1      1          0        0   155
#> [368,]  73   6      1          0        1   270
#> [369,]  72   8      1          1        1   150
#> [370,]  71   2      1          0        1   180
#> [371,]  70  13      1          9        0   100
#> [372,]  72   6      0        180        1   130
#> [373,]  74   8      1          0        1    85
#> [374,]  73   4      0        180        1   154
#> [375,]  69   2      1          0        1   110
#> [376,]  71  15      1         11        0   165
#> [377,]  74  20      0         20        1   180
#> [378,]  68   9      0        180        1   120
#> [379,]  71  20      1         10        0   140
#> [380,]  74   0      1          0        1    90
#> [381,]  73   3      1          0        1   136
#> [382,]  70   5      1          0        1   190
#> [383,]  71   8      1          7        0   149
#> [384,]  71   3      1          2        1   190
#> [385,]  70  26      1         11        1   120
#> [386,]  73   4      0         58        1   160
#> [387,]  72   5      1          3        1   160
#> [388,]  70   3      0        180        1   154
#> [389,]  71  13      1          8        0   121
#> [390,]  69   2      1          1        1    80
#> [391,]  70   4      1          0        1   140
#> [392,]  74   7      1          0        1   117
#> [393,]  72  10      1          8        1   153
#> [394,]  69   7      0        180        1   144
#> [395,]  72  15      1         13        0   156
#> [396,]  71  10      1          9        1   120
#> [397,]  75   1      0          1        0   133
#> [398,]  75   2      1          1        0   145
#> [399,]  73  10      1          9        1   146
#> [400,]  72  10      1          9        1   160
#> [401,]  73  10      1         10        1   120
#> [402,]  74  15      1          9        1   179
#> [403,]  71   2      0         10        1   112
#> [404,]  75   9      1          7        0   140
#> [405,]  75  13      1          1        1   130
#> [406,]  71  11      1          8        0   110
#> [407,]  71   4      0          4        0   134
#> [408,]  73  10      1          8        0   120
#> [409,]  72   1      1          1        0   168
#> [410,]  72  11      0         11        1   140
#> [411,]  70   3      0          3        0   150
#> [412,]  73   5      1          3        1   112
#> [413,]  72   2      0        180        0   120
#> [414,]  75   1      0        180        1   140
#> [415,]  72   4      1          0        1   197
#> [416,]  73   5      0        180        0   126
#> [417,]  74  34      1          8        1   233
#> [418,]  76   3      1          0        1   120
#> [419,]  71  32      1         12        1   107
#> [420,]  72   5      0        180        0   154
#> [421,]  72   3      0        180        0   160
#> [422,]  76   5      0          5        1   130
#> [423,]  77  11      0         11        1   150
#> [424,]  75   3      1          1        0   180
#> [425,]  72   7      1          2        0   142
#> [426,]  73  15      0         15        1   160
#> [427,]  71  16      0        180        0   140
#> [428,]  73  10      1         10        0   124
#> [429,]  74   7      0        180        1   150
#> [430,]  74   2      1          1        0   140
#> [431,]  74  19      1          4        1   200
#> [432,]  73   6      0          6        1   114
#> [433,]  75  23      1         14        1   110
#> [434,]  72   4      1          3        0   160
#> [435,]  73   4      1          3        1   125
#> [436,]  75   7      0          7        0   190
#> [437,]  75   0      0          0        1   130
#> [438,]  73  13      1         11        0   195
#> [439,]  75  12      0         12        1   160
#> [440,]  75   4      1          2        1   188
#> [441,]  74   6      0        180        0   160
#> [442,]  76   4      0          4        1   155
#> [443,]  75   1      0          1        1   125
#> [444,]  74   2      0        180        0   111
#> [445,]  73   1      0         52        1   105
#> [446,]  73   0      0        180        0   156
#> [447,]  76  44      1         10        0   105
#> [448,]  76   5      0        180        0   185
#> [449,]  74   8      1          8        1   170
#> [450,]  73  33      1         12        1   175
#> [451,]  77   5      1          0        0   123
#> [452,]  77  12      1          9        1   100
#> [453,]  73  10      1          9        0   146
#> [454,]  77  12      0        180        0   130
#> [455,]  77   1      1          0        1    90
#> [456,]  76  12      1         11        1   120
#> [457,]  75   3      1          1        1   171
#> [458,]  74   9      1          8        0   200
#> [459,]  75   6      0        180        0   150
#> [460,]  79  10      1          8        0   190
#> [461,]  74   2      1          0        1   130
#> [462,]  78  18      0         18        1   144
#> [463,]  77   3      0        180        0   110
#> [464,]  76  29      0         47        0    90
#> [465,]  73  11      1          2        1   110
#> [466,]  74   2      0        180        0   100
#> [467,]  78   7      0          7        1   133
#> [468,]  78   8      1          6        1   110
#> [469,]  74   7      0          7        0   161
#> [470,]  79   6      0        180        0   170
#> [471,]  78   0      0        180        1   212
#> [472,]  78  13      1          5        0   130
#> [473,]  75   5      0        119        1   150
#> [474,]  75  12      1          1        1   120
#> [475,]  78  15      0        180        1   270
#> [476,]  80   8      0          8        1   120
#> [477,]  75  13      1          6        0   150
#> [478,]  76   1      0          1        1    83
#> [479,]  79   4      0         80        0   145
#> [480,]  78  12      1          9        0   150
#> [481,]  78   2      1          1        0   130
#> [482,]  77   2      1          0        1   143
#> [483,]  78  10      0        180        1   130
#> [484,]  75  11      1          4        0   162
#> [485,]  76   7      0         29        1   150
#> [486,]  77  24      0         24        1   160
#> [487,]  79   8      0         32        1   120
#> [488,]  80   9      0         23        1   128
#> [489,]  78   6      1          0        1   240
#> [490,]  78  11      1          1        1   140
#> [491,]  79  11      0        180        0   160
#> [492,]  79   2      1          0        1   121
#> [493,]  78  14      1          0        1   140
#> [494,]  78  11      1          8        1   118
#> [495,]  76   4      0          4        1   160
#> [496,]  79   4      0          4        1   125
#> [497,]  76  10      1          8        0   180
#> [498,]  76  12      1         10        1   127
#> [499,]  75   2      1          1        1   204
#> [500,]  78  11      0        180        1   135
#> [501,]  77  31      1          3        1   161
#> [502,]  76   1      0          1        1    90
#> [503,]  79   3      0          3        0   120
#> [504,]  77   7      0        180        1   170
#> [505,]  77   6      0          6        1   144
#> [506,]  79   4      1          0        1   120
#> [507,]  81   1      0        180        0   120
#> [508,]  80  15      1         12        1   150
#> [509,]  77   9      1          4        0   141
#> [510,]  82   5      0          8        1   120
#> [511,]  78   4      0         59        1   112
#> [512,]  76   7      0        161        0   151
#> [513,]  79  10      0         10        1   120
#> [514,]  80  15      1          0        1    90
#> [515,]  81   4      1          2        1   126
#> [516,]  79  28      0        164        0   100
#> [517,]  80   9      0        118        1   186
#> [518,]  80   6      0        173        1   160
#> [519,]  78  32      0        180        1   130
#> [520,]  81   3      0        180        0   184
#> [521,]  78   7      0          7        1   147
#> [522,]  77  13      1          0        1   190
#> [523,]  78  15      0         15        0   165
#> [524,]  78   4      0        180        0   175
#> [525,]  79   3      0          3        1   101
#> [526,]  78  26      1          5        0   194
#> [527,]  81   4      1          1        1   104
#> [528,]  78  20      1          0        1   109
#> [529,]  80   1      0          1        0   100
#> [530,]  78   3      1          1        1   152
#> [531,]  80   2      1          1        0   168
#> [532,]  80   6      1          0        1   119
#> [533,]  78   2      0        180        0   148
#> [534,]  80   5      0          5        1   130
#> [535,]  82   1      1          0        1    82
#> [536,]  79  10      0        180        1   150
#> [537,]  77   4      0        180        1    98
#> [538,]  81   1      0        108        0   129
#> [539,]  78  12      0        180        0   134
#> [540,]  79   1      0        125        0   193
#> [541,]  82  21      1          2        0   155
#> [542,]  80   6      0          6        1   110
#> [543,]  83   9      1          5        1   170
#> [544,]  82   5      0        180        0   110
#> [545,]  79   7      1          6        0   130
#> [546,]  83   4      0        103        0    97
#> [547,]  81   5      0        177        0    41
#> [548,]  80  11      1          8        0   170
#> [549,]  79   4      0          4        1   183
#> [550,]  78   9      1          4        1   120
#> [551,]  79   1      0        180        1   170
#> [552,]  81  15      0        180        1   140
#> [553,]  80   7      1          0        1   146
#> [554,]  81  20      1          9        0   170
#> [555,]  81  16      0         16        1   110
#> [556,]  80   6      1          0        1   150
#> [557,]  81   8      0        180        0   146
#> [558,]  79   7      0        177        0   197
#> [559,]  79   0      1          0        1    96
#> [560,]  81   2      1          1        0   198
#> [561,]  83   2      0          2        1   155
#> [562,]  82   6      0        128        1   100
#> [563,]  80   3      1          1        1   230
#> [564,]  82  23      1          0        0   110
#> [565,]  84   5      0        180        1   203
#> [566,]  84   4      0          4        1    85
#> [567,]  81   1      0          1        1   150
#> [568,]  83   3      0        180        0   174
#> [569,]  81   4      0         90        1   138
#> [570,]  79   9      1          8        0   150
#> [571,]  85   3      1          2        1   160
#> [572,]  80   2      1          0        1   130
#> [573,]  79   4      0          4        1    60
#> [574,]  80   6      0         71        1   189
#> [575,]  83   1      0          1        1   100
#> [576,]  82  19      0         19        0   120
#> [577,]  83   9      0        180        0   198
#> [578,]  79  14      1          0        0   110
#> [579,]  85   9      1          6        1   160
#> [580,]  81   4      0          4        0   175
#> [581,]  84  15      1         13        1   110
#> [582,]  81   1      0          1        1   145
#> [583,]  82  16      1          8        0   103
#> [584,]  82   5      1          0        1   146
#> [585,]  83  12      1          2        1   170
#> [586,]  81  19      1         14        0   120
#> [587,]  82   3      1          2        0   130
#> [588,]  82  15      1          0        0   183
#> [589,]  83   7      0        126        0   135
#> [590,]  86   8      0          8        1   132
#> [591,]  81  16      1          9        0   180
#> [592,]  86   3      0          3        1   140
#> [593,]  81  13      0        180        0   152
#> [594,]  85   3      0          3        1   118
#> [595,]  81   4      0        180        0   160
#> [596,]  82   1      0        180        1   193
#> [597,]  86  12      1          0        1   132
#> [598,]  82  14      1         11        1   103
#> [599,]  86   6      1          0        1   140
#> [600,]  84  16      0         70        1   150
#> [601,]  83  10      1          0        1   190
#> [602,]  86   2      0        180        1   169
#> [603,]  88  14      1          3        1   130
#> [604,]  84   3      0          3        1   121
#> [605,]  84   7      1          2        0   148
#> [606,]  87   2      0        180        0   113
#> [607,]  84   9      0         92        1   110
#> [608,]  84   3      0        180        1   170
#> [609,]  82   4      0          4        0   130
#> [610,]  85   3      0          3        1   113
#> [611,]  86   6      0          6        1   117
#> [612,]  84  13      0         62        1   100
#> [613,]  86   6      1          1        0   112
#> [614,]  88   4      0          4        0   100
#> [615,]  83  20      1          3        1   150
#> [616,]  85  22      0         22        1   184
#> [617,]  83   9      0         65        1   150
#> [618,]  86   9      1          7        1   142
#> [619,]  88   3      0        115        0   110
#> [620,]  88   2      0        180        1    68
#> [621,]  83   3      0          3        1   130
#> [622,]  87   8      0          8        1   157
#> [623,]  86  15      1          8        1   109
#> [624,]  89   4      0          4        1   153
#> [625,]  89   5      0        119        1   140
#> [626,]  87   6      0        180        1   110
#> [627,]  84   8      0        180        1   119
#> [628,]  85   8      0          8        1   136
#> [629,]  89  10      0         46        1   170
#> [630,]  90  14      0         14        1   100
#> [631,]  88   1      0          1        0   135
#> [632,]  91   8      0          8        0   100
#> [633,]  87   2      0        180        0   160
#> [634,]  87   6      1          0        0   125
#> [635,]  86   3      1          0        1    80
#> [636,]  88   7      0         24        0   119
#> [637,]  88   8      0         50        1   154
#> [638,]  90  11      1         10        1   186
#> [639,]  86  10      0        180        1   137
#> [640,]  86   9      1          7        0   130
#> [641,]  91   1      0          1        1    74
#> [642,]  87  43      0        178        1   130
#> [643,]  87   5      0         36        1   150
#> [644,]  90   5      1          0        1   125
#> [645,]  91   3      0         33        1   137
#> [646,]  88   5      0        158        0   100
#> [647,]  87   7      0         74        1   105
#> [648,]  89  12      1          0        1   130
#> [649,]  89   2      0        168        0   118
#> [650,]  91   5      0        169        1   176
#> [651,]  91   0      0          0        0     0
#> [652,]  89  14      0        180        1    84
#> [653,]  90  18      0        180        0   188
#> [654,]  91   4      1          0        1   120
#> [655,]  90  19      1         11        1   129
#> [656,]  94   6      0         50        0    78
#> [657,]  90   1      0          1        1   118
#> [658,]  91   2      0          2        1   116
#> [659,]  93   8      0        179        1   110
#> [660,]  94   8      0          8        1   142
#> [661,]  92   4      0         76        1   149
#> [662,]  91   1      0        180        0   158
#> [663,]  90  16      0         16        1   106
#> [664,]  90   3      0         67        0   162
#> [665,]  96   3      0         12        1    97
#> [666,]  94   3      0         26        1   144
#> [667,]  91  12      0         53        1   212
#> [668,]  91   7      0          7        0   135
#> [669,]  93   0      1          0        1   122
#> [670,]  96   3      1          0        1   104
#> 
#> $y
#>   [1] 180.0+   5.0+   2.0+   5.0+ 180.0+   2.0+ 180.0+ 115.0    5.0+ 180.0+
#>  [11] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 180.0+ 180.0+ 180.0+
#>  [21]   3.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0+
#>  [31] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [41] 180.0+   6.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [51] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  10.0+ 172.0+   7.0 
#>  [61] 180.0+ 180.0+   2.0    1.0  180.0+ 179.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [71] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   7.0   88.0+ 180.0+   4.0+ 180.0+
#>  [81] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  85.0  180.0+
#>  [91] 166.0+  99.0  180.0+ 180.0+ 152.0+ 180.0+ 180.0+ 180.0+ 171.0+ 174.0+
#> [101]  28.0    6.0+   9.0+ 180.0+ 175.0+ 180.0+   2.0    7.0+ 180.0+ 180.0+
#> [111] 180.0+ 180.0+ 180.0+  16.0+  16.0   15.0+ 180.0+ 180.0+  12.0+ 134.0+
#> [121]   3.0+ 180.0+ 180.0+ 180.0+ 140.0  180.0+   1.0  165.0  180.0+ 180.0+
#> [131] 180.0+ 180.0+ 180.0+   8.0+ 180.0+   0.5  180.0+ 180.0+ 171.0+  15.0 
#> [141] 166.0+ 180.0+   4.0+ 180.0+ 180.0+   5.0+ 180.0+   4.0+ 180.0+   9.0+
#> [151] 180.0+   2.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0  180.0+ 180.0+ 180.0+
#> [161]  64.0  180.0+ 180.0+   0.5  180.0+ 180.0+   3.0    1.0  180.0+ 180.0+
#> [171]  10.0+ 180.0+  17.0   45.0    3.0+  19.0  180.0+ 180.0+   9.0+ 180.0+
#> [181] 172.0+ 180.0+   8.0  180.0+  15.0  180.0+ 180.0+ 180.0+  77.0    8.0+
#> [191] 180.0+ 180.0+  94.0  180.0+ 169.0    7.0    7.0+   6.0  180.0+   3.0+
#> [201] 180.0+  28.0    5.0   18.0  180.0+ 180.0+ 180.0+   1.0  180.0+ 180.0+
#> [211] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   9.0    7.0+
#> [221]  22.0  180.0+  84.0  180.0+ 180.0+   1.0  180.0+   3.0+ 167.0  180.0+
#> [231] 180.0+  12.0  180.0+ 180.0+  36.0    3.0+  88.0  180.0+  12.0  180.0+
#> [241] 180.0+   0.5  180.0+  12.0+ 180.0+ 180.0+ 180.0+  12.0    9.0  180.0+
#> [251]  11.0+   2.0+  18.0+ 180.0+ 180.0+ 180.0+ 180.0+   3.0+   2.0+ 103.0 
#> [261]  15.0    5.0+ 180.0+ 166.0+  14.0+   0.5+ 180.0+   3.0+ 180.0+ 175.0+
#> [271] 180.0+   8.0    5.0   16.0    1.0  180.0+ 123.0+   1.0+ 180.0+  79.0 
#> [281] 180.0+   4.0+  15.0  180.0+ 180.0+ 180.0+ 174.0+ 180.0+   8.0+ 175.0 
#> [291]  10.0  180.0+ 180.0+ 180.0+   6.0  180.0+  99.0  179.0+ 180.0+ 180.0+
#> [301]   0.5   18.0  152.0+  93.0  180.0+   1.0   18.0+ 101.0    4.0    5.0 
#> [311] 180.0+ 180.0+ 180.0+ 180.0+ 174.0+ 180.0+ 180.0+ 180.0+ 180.0+   7.0+
#> [321]   2.0   45.0    5.0+ 103.0  180.0+  36.0    5.0+ 180.0+   7.0    8.0+
#> [331] 180.0+   2.0+  18.0  172.0+ 180.0+ 180.0+   7.0  180.0+   7.0    8.0+
#> [341]  13.0+ 180.0+ 180.0+  51.0   19.0  180.0+   1.0  132.0  180.0+ 124.0 
#> [351]   9.0  180.0+ 180.0+  12.0  180.0+ 152.0  180.0+   2.0   76.0  173.0+
#> [361] 180.0+ 180.0+  28.0  180.0+ 180.0+  16.0+ 180.0+   6.0  180.0+ 180.0+
#> [371]  13.0+ 180.0+ 180.0+ 180.0+   2.0  180.0+  20.0  180.0+  20.0    0.5 
#> [381] 180.0+ 180.0+   8.0    3.0  180.0+  58.0  180.0+ 180.0+ 175.0    2.0 
#> [391] 180.0+ 180.0+  10.0+ 180.0+ 180.0+ 179.0+   1.0  180.0+ 180.0+ 159.0 
#> [401]  15.0  180.0+  10.0  180.0+  13.0  180.0+   4.0+  10.0    1.0   11.0 
#> [411]   3.0+   5.0  180.0+ 180.0+ 180.0+ 180.0+  34.0  180.0+ 177.0+ 180.0+
#> [421] 180.0+   5.0   11.0+ 180.0+   7.0   15.0+ 180.0+  10.0  180.0+ 180.0+
#> [431] 180.0+   6.0  180.0+ 180.0+ 180.0+   7.0    0.5  180.0+  12.0   46.0 
#> [441] 180.0+   4.0    1.0  180.0+  52.0  180.0+ 180.0+ 180.0+   8.0   33.0 
#> [451]   5.0  180.0+ 180.0+ 180.0+   1.0   12.0    3.0  168.0+ 180.0+ 180.0+
#> [461] 176.0+  18.0  180.0+  47.0   11.0  180.0+   7.0    8.0+   7.0  180.0+
#> [471] 180.0+ 172.0  119.0   12.0  180.0+   8.0  180.0+   1.0   80.0  180.0+
#> [481] 180.0+   2.0  180.0+ 152.0+  29.0   24.0   32.0   23.0  180.0+ 180.0+
#> [491] 180.0+ 180.0+ 180.0+  11.0    4.0    4.0   10.0+ 180.0+   2.0+ 180.0+
#> [501] 171.0    1.0    3.0  180.0+   6.0  138.0  180.0+ 180.0+  71.0    8.0 
#> [511]  59.0  161.0   10.0+ 180.0+  93.0  164.0  118.0  173.0  180.0+ 180.0+
#> [521]   7.0+  22.0   15.0+ 180.0+   3.0  171.0+  71.0   20.0+   1.0    3.0+
#> [531]  10.0    6.0  180.0+   5.0    1.0  180.0+ 180.0+ 108.0  180.0+ 125.0 
#> [541] 180.0+   6.0    9.0+ 180.0+ 180.0+ 103.0  177.0+ 169.0    4.0  180.0+
#> [551] 180.0+ 180.0+   7.0+  20.0   16.0  180.0+ 180.0+ 177.0+   0.5  180.0+
#> [561]   2.0  128.0    3.0+  62.0  180.0+   4.0    1.0  180.0+  90.0  180.0+
#> [571] 180.0+ 180.0+   4.0   71.0    1.0   19.0  180.0+ 180.0+ 180.0+   4.0+
#> [581] 180.0+   1.0   16.0+   5.0+  77.0  180.0+   3.0   83.0  126.0    8.0 
#> [591] 180.0+   3.0  180.0+   3.0+ 180.0+ 180.0+ 180.0+ 174.0    6.0   70.0 
#> [601] 180.0+ 180.0+  14.0    3.0  180.0+ 180.0+  92.0  180.0+   4.0    3.0+
#> [611]   6.0+  62.0    6.0+   4.0+  20.0   22.0   65.0   11.0  115.0  180.0+
#> [621]   3.0+   8.0+ 180.0+   4.0  119.0  180.0+ 180.0+   8.0   46.0   14.0 
#> [631]   1.0+   8.0  180.0+  25.0    3.0   24.0   50.0   11.0  180.0+ 180.0+
#> [641]   1.0  178.0+  36.0   89.0   33.0  158.0   74.0  180.0+ 168.0  169.0 
#> [651]   0.5  180.0+ 180.0+   4.0  180.0+  50.0    1.0+   2.0  179.0+   8.0+
#> [661]  76.0  180.0+  16.0   67.0   12.0   26.0   53.0    7.0+   0.5    3.0 
#> 
#> $weights
#> NULL
#> 
#> $offset
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8408723