Cross-Validated GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::cv.glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.cv.glmnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.cv.glmnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.cv_glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

foldid	untyped	NULL		-

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nfolds	integer	10		\([3, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	TRUE	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	lambda.1se		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

standardize.response	logical	FALSE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.gaussian	character	-	covariance, naive	-

type.logistic	character	Newton	Newton, modified.Newton	-

type.measure	character	deviance	deviance, C	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.coxnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvCVGlmnet

Methods

Public methods

LearnerSurvCVGlmnet$new()

LearnerSurvCVGlmnet$selected_features()

LearnerSurvCVGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvCVGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvCVGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvCVGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("surv.cv_glmnet")
print(learner)
#> 
#> ── <LearnerSurvCVGlmnet> (surv.cv_glmnet): Regularized Generalized Linear Model 
#> • Model: -
#> • Parameters: use_pred_offset=TRUE
#> • Packages: mlr3, mlr3proba, mlr3extralearners, and glmnet
#> • Predict Types: [crank], distr, and lp
#> • Feature Types: logical, integer, and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: offset, selected_features, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("grace")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#> Measure: Partial Likelihood Deviance 
#> 
#>      Lambda Index Measure     SE Nonzero
#> min 0.00337    45   2.708 0.1424       6
#> 1se 0.07260    12   2.842 0.1543       3
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  28   9      0        180        1   107
#>   [2,]  32   5      1          0        1   121
#>   [3,]  35   5      1          2        0   172
#>   [4,]  35  10      1          9        0   106
#>   [5,]  35   2      0        180        0   121
#>   [6,]  35   2      1          1        1   112
#>   [7,]  37   9      0        180        1   151
#>   [8,]  35   0      0        180        1   119
#>   [9,]  38  12      1          8        1   120
#>  [10,]  36   5      1          0        1   115
#>  [11,]  33   6      1          1        1   115
#>  [12,]  38  12      1         11        1    92
#>  [13,]  40  12      1          9        0   153
#>  [14,]  42   3      1          1        1   130
#>  [15,]  37   1      1          0        1   146
#>  [16,]  42   2      0          2        0   140
#>  [17,]  40   6      0        180        1   138
#>  [18,]  40  11      1         10        1   120
#>  [19,]  42   2      0        180        0   100
#>  [20,]  43   3      1          0        1   100
#>  [21,]  41   2      1          1        0   166
#>  [22,]  40   1      1          0        1   145
#>  [23,]  43   4      1          0        1   130
#>  [24,]  42   4      0        180        0   162
#>  [25,]  40   3      1          1        0   170
#>  [26,]  42  12      1         10        1   170
#>  [27,]  43   2      1          1        1   116
#>  [28,]  42   2      0        180        1   124
#>  [29,]  44   5      1          1        0   170
#>  [30,]  45   3      0        180        1   154
#>  [31,]  41  10      1          8        0   150
#>  [32,]  44   3      0        180        0   141
#>  [33,]  41  13      1          1        0   140
#>  [34,]  45   9      1          7        0   110
#>  [35,]  45   6      0        180        1   170
#>  [36,]  41   5      1          4        1   141
#>  [37,]  44   2      1          1        1   150
#>  [38,]  43   2      0        180        1   140
#>  [39,]  45   2      0        180        1   140
#>  [40,]  46  15      0        180        0   120
#>  [41,]  46   2      1          1        0   126
#>  [42,]  47   4      1          3        0   118
#>  [43,]  45   3      0        150        0   130
#>  [44,]  44   3      1          0        1   180
#>  [45,]  46   7      1          2        0   166
#>  [46,]  43  29      0        180        1   180
#>  [47,]  45   4      1          0        0   124
#>  [48,]  46  13      1         10        0   100
#>  [49,]  44   0      1          0        1    96
#>  [50,]  47   4      1          3        1   160
#>  [51,]  43   3      1          0        1   124
#>  [52,]  45   8      1          0        1   117
#>  [53,]  46   2      1          1        1   122
#>  [54,]  46   6      1          0        1   100
#>  [55,]  44   4      1          0        1   114
#>  [56,]  47   2      0        180        0   108
#>  [57,]  44   9      1          8        1   135
#>  [58,]  45   5      0        180        1   190
#>  [59,]  46   5      1          3        0   130
#>  [60,]  46   4      0        180        1   121
#>  [61,]  44   2      0        180        0   142
#>  [62,]  47   3      1          1        1   120
#>  [63,]  48   3      0        180        0   154
#>  [64,]  48  12      1         11        0   200
#>  [65,]  47   5      1          3        1   130
#>  [66,]  47   9      1          6        0   170
#>  [67,]  49   4      0        180        0   117
#>  [68,]  50   1      1          0        1   129
#>  [69,]  50   6      1          2        1   140
#>  [70,]  46   3      1          1        1   140
#>  [71,]  50   7      0        180        1   110
#>  [72,]  49   2      0          2        0   105
#>  [73,]  51   1      0          1        1   145
#>  [74,]  49  15      1         11        1   160
#>  [75,]  47   2      0        180        0   150
#>  [76,]  46   6      1          0        1   156
#>  [77,]  52   2      0        180        1   170
#>  [78,]  50   7      1          0        1    92
#>  [79,]  51   3      1          2        0   113
#>  [80,]  50   1      1          0        0   150
#>  [81,]  50   9      0        180        0   130
#>  [82,]  49   7      1          4        1    90
#>  [83,]  47   6      0        180        1   162
#>  [84,]  51   8      0        180        1   140
#>  [85,]  52   2      0        180        0   155
#>  [86,]  46   1      1          1        0   145
#>  [87,]  50   4      1          1        0   150
#>  [88,]  48   7      1          0        1   110
#>  [89,]  53   8      0         36        1   160
#>  [90,]  47   2      1          1        0   110
#>  [91,]  52   4      1          4        0   152
#>  [92,]  53   5      0        180        1   140
#>  [93,]  54  17      1         12        1   102
#>  [94,]  53   5      0         77        0   159
#>  [95,]  53   7      1          0        0   199
#>  [96,]  54   6      1          3        0   129
#>  [97,]  50  14      1         13        0   170
#>  [98,]  48   3      1          2        0   150
#>  [99,]  51  25      1          1        0   202
#> [100,]  49   5      1          2        1   150
#> [101,]  53   4      0          4        0   140
#> [102,]  48  11      1         10        0   120
#> [103,]  53   4      1          0        1   156
#> [104,]  51  13      0         99        1   160
#> [105,]  54  23      1         10        0   131
#> [106,]  52   7      1          2        0   154
#> [107,]  54   9      1          1        0   130
#> [108,]  55   4      1          2        0   150
#> [109,]  52   4      0        180        1   180
#> [110,]  51  13      1         11        0   145
#> [111,]  50   5      1          4        1   150
#> [112,]  50   3      0        174        1   153
#> [113,]  55  28      1         13        1   160
#> [114,]  49   6      1          0        1   130
#> [115,]  50   7      1          1        0   156
#> [116,]  53   8      1          0        1   130
#> [117,]  50   7      1          0        1   127
#> [118,]  52   5      0        175        1   117
#> [119,]  55   2      0          2        0   145
#> [120,]  54   1      0        180        0   162
#> [121,]  54   7      1          0        1   100
#> [122,]  56   3      0        180        1   193
#> [123,]  55   5      1          4        1   120
#> [124,]  52   8      0        180        0   119
#> [125,]  53  18      1          9        1   150
#> [126,]  54   3      0        180        1   180
#> [127,]  55   6      0        180        0   170
#> [128,]  53  10      1          9        0   172
#> [129,]  52  16      1         14        0   170
#> [130,]  53   4      0        180        1   150
#> [131,]  55   6      0        180        1   100
#> [132,]  55   6      1          5        1   138
#> [133,]  54  12      1          0        1   190
#> [134,]  54   3      0        180        0   128
#> [135,]  55   1      0          2        0   130
#> [136,]  54   7      1          2        0   129
#> [137,]  54   2      1          1        1   176
#> [138,]  57   5      1          3        1   138
#> [139,]  56   4      1          0        1   140
#> [140,]  52   2      0        180        0   140
#> [141,]  55  11      1          7        0   104
#> [142,]  52  15      1         14        0   130
#> [143,]  53   3      1          0        1   200
#> [144,]  58   8      0          8        1   130
#> [145,]  54   5      0        180        1   108
#> [146,]  55   3      1          1        1   156
#> [147,]  57   0      0          0        1   150
#> [148,]  59   3      1          1        0   172
#> [149,]  58   6      1          0        1    90
#> [150,]  53  15      1         10        1   130
#> [151,]  54  17      1          8        1   227
#> [152,]  55   9      1          2        1   147
#> [153,]  55  13      0        166        1   140
#> [154,]  54  23      1          8        0   120
#> [155,]  53   4      0        147        1   145
#> [156,]  59  15      1         10        0   140
#> [157,]  58   9      1          0        1   180
#> [158,]  58   1      1          1        1   200
#> [159,]  55   5      1          0        0   140
#> [160,]  56   7      1          5        1   120
#> [161,]  55   2      0          2        0   106
#> [162,]  59   9      1          1        1   125
#> [163,]  57   1      0        180        0   148
#> [164,]  60  11      1          9        0   106
#> [165,]  60   5      1          1        0   138
#> [166,]  57   5      0        180        1   130
#> [167,]  58  11      1          9        1   124
#> [168,]  55   5      1          0        1   160
#> [169,]  57  10      1          9        0   103
#> [170,]  59   6      1          0        1   140
#> [171,]  59   5      0        180        1   155
#> [172,]  59   4      1          0        1   152
#> [173,]  58  26      1          0        1   189
#> [174,]  58   4      1          3        0   120
#> [175,]  58   8      0        161        1   140
#> [176,]  58  14      1          6        0   190
#> [177,]  61   4      1          3        0   151
#> [178,]  61   9      1          8        0   150
#> [179,]  61   3      1          2        1   102
#> [180,]  58   1      0          1        1   100
#> [181,]  61  20      1         13        0   130
#> [182,]  57  13      1         10        0   110
#> [183,]  57   2      1          0        1   116
#> [184,]  57  11      0        180        1   150
#> [185,]  56  14      0         45        0   130
#> [186,]  57   3      1          2        0   120
#> [187,]  58  19      1         13        1   140
#> [188,]  56  13      1          6        1   158
#> [189,]  56  18      1         11        1   165
#> [190,]  59   9      1          0        1    80
#> [191,]  55   4      1          3        1   160
#> [192,]  58  11      0        172        1   135
#> [193,]  60  12      1          0        1   114
#> [194,]  56   8      1          8        0   120
#> [195,]  57   1      0          1        0   126
#> [196,]  57  15      1         13        1   110
#> [197,]  59   5      1          2        0   182
#> [198,]  59  10      0        180        0   160
#> [199,]  58   8      1          5        0   152
#> [200,]  62  10      1          0        1   153
#> [201,]  62   7      1          2        1   180
#> [202,]  61  18      0        170        0   140
#> [203,]  58   8      1          3        1   150
#> [204,]  57   7      0        169        0   180
#> [205,]  60   7      0          7        0   147
#> [206,]  61   6      0          6        0   134
#> [207,]  62   4      1          0        0   160
#> [208,]  58   3      1          0        1   146
#> [209,]  58   2      0         30        0   202
#> [210,]  59   1      0        180        0   155
#> [211,]  61  13      0         13        0   120
#> [212,]  61   5      0          5        1   160
#> [213,]  58  11      1          9        0   179
#> [214,]  62  17      1         10        1   180
#> [215,]  62   1      1          0        1   172
#> [216,]  58   7      0        180        1   150
#> [217,]  63   3      1          1        0   180
#> [218,]  63   4      1          3        0   222
#> [219,]  62   3      0        180        1   105
#> [220,]  63   4      0        180        1   190
#> [221,]  64   4      0        180        0   130
#> [222,]  59   8      0        180        1   140
#> [223,]  61   9      1          9        1   150
#> [224,]  58   9      1          9        0   110
#> [225,]  62   7      0          7        0   150
#> [226,]  59   1      0         22        1   162
#> [227,]  59   4      0        180        0   196
#> [228,]  60   7      1          5        1   141
#> [229,]  60   7      0          7        0   140
#> [230,]  59   5      1          1        0   148
#> [231,]  65  13      0        180        1   100
#> [232,]  63   1      0          1        0   130
#> [233,]  62   6      0        180        0   170
#> [234,]  61  15      1         13        0   170
#> [235,]  59   4      0          4        0   149
#> [236,]  60   3      0          3        0   168
#> [237,]  62   6      0          6        0   120
#> [238,]  63  12      1         10        0   200
#> [239,]  59  10      0        180        1   130
#> [240,]  60   8      0         17        1   130
#> [241,]  64  12      1         11        0   160
#> [242,]  64   6      1          0        1   140
#> [243,]  63  14      1          9        0   123
#> [244,]  65  36      1         11        0   140
#> [245,]  63   4      1          3        0   162
#> [246,]  66   3      1          1        0   127
#> [247,]  64  32      1          9        1   160
#> [248,]  63  12      1          9        0   114
#> [249,]  63   7      0        180        0   120
#> [250,]  64   0      0          0        1    90
#> [251,]  60   6      0        180        0   130
#> [252,]  64  21      1         10        0   190
#> [253,]  64   9      0        180        0   150
#> [254,]  65   3      0        180        1   190
#> [255,]  63  16      1          7        1   110
#> [256,]  64   7      0        180        1   120
#> [257,]  66   6      1          1        1   130
#> [258,]  63  12      0         12        1   150
#> [259,]  62   3      1          1        1   199
#> [260,]  65   6      0          9        0   112
#> [261,]  63   2      1          1        0   180
#> [262,]  67  11      0         11        1   100
#> [263,]  64   2      0          2        0   201
#> [264,]  66  16      1         11        1   169
#> [265,]  62   9      0        180        0   145
#> [266,]  61  14      1          5        0   140
#> [267,]  61  15      1         10        0   130
#> [268,]  63   9      1          8        1   160
#> [269,]  63   3      1          2        0   120
#> [270,]  64  19      1          8        1   160
#> [271,]  65  15      1         11        1   160
#> [272,]  68   5      1          4        1   150
#> [273,]  64  13      1         12        1   150
#> [274,]  64   6      1          0        1   125
#> [275,]  66   7      1          0        1   115
#> [276,]  66  13      1          0        0   118
#> [277,]  65   3      0          3        0   105
#> [278,]  64   0      0          0        1   148
#> [279,]  67   4      1          3        0   130
#> [280,]  66   6      1          0        1   140
#> [281,]  65   2      1          1        1   170
#> [282,]  68   1      0        180        1   166
#> [283,]  63   7      1          0        0   162
#> [284,]  68   5      0          5        1    90
#> [285,]  63  10      0         16        1   160
#> [286,]  66  14      0        180        0   130
#> [287,]  64   1      0          1        1   120
#> [288,]  68  18      0        180        1   260
#> [289,]  65  18      1          3        0   120
#> [290,]  63   1      1          0        1   155
#> [291,]  63  10      0         18        1   130
#> [292,]  67  11      0         11        0   150
#> [293,]  68  11      0        180        0   160
#> [294,]  68  14      0         79        0   172
#> [295,]  66  15      1         13        1   160
#> [296,]  63   2      0        180        0   150
#> [297,]  69   6      0        180        1   100
#> [298,]  63   8      0        180        1   120
#> [299,]  68  14      1         13        1   140
#> [300,]  65   8      1          0        1    90
#> [301,]  67   1      0        180        1   160
#> [302,]  68  10      1         10        1   150
#> [303,]  65   1      1          0        0   133
#> [304,]  67   7      1          4        1   130
#> [305,]  63   2      1          0        0    99
#> [306,]  67   2      0        180        0   184
#> [307,]  66  19      1         12        1   150
#> [308,]  67  12      1         12        0   160
#> [309,]  69   6      0         99        1   140
#> [310,]  64   4      0        179        0   160
#> [311,]  70  15      1         12        1   132
#> [312,]  64   4      0        180        1   140
#> [313,]  64   0      1          0        1   118
#> [314,]  67   2      0         18        0   131
#> [315,]  66   7      1          5        1   131
#> [316,]  66   4      0        180        0   177
#> [317,]  68   4      1          0        1   160
#> [318,]  69   4      1          3        1   150
#> [319,]  65  13      1         12        1   130
#> [320,]  69  17      1         10        0   140
#> [321,]  69   8      0         93        0   140
#> [322,]  64  21      0         21        1   155
#> [323,]  66   6      0        180        0   140
#> [324,]  65   1      0          1        1   120
#> [325,]  68  18      1          0        1   160
#> [326,]  71   3      0          5        0   112
#> [327,]  70   7      1          0        1   190
#> [328,]  68   7      0        150        0   210
#> [329,]  71  20      1          0        1   160
#> [330,]  66   9      1          3        1   151
#> [331,]  70   4      1          0        1   180
#> [332,]  70  14      0        171        0   166
#> [333,]  66   4      0        180        0   130
#> [334,]  67   6      1          4        0   130
#> [335,]  65   2      0        180        0   130
#> [336,]  68   7      1          0        1   150
#> [337,]  69   3      1          2        0   151
#> [338,]  65  14      1         13        1   150
#> [339,]  69   8      0        180        1   180
#> [340,]  66   2      0          2        1   228
#> [341,]  69   5      0          5        1   142
#> [342,]  71   3      0        103        0   133
#> [343,]  67   5      0          5        0   130
#> [344,]  69   8      1          5        1   195
#> [345,]  72   7      0          7        1   110
#> [346,]  67  14      0        172        1   140
#> [347,]  67   7      1          4        0   122
#> [348,]  69   4      1          3        0   132
#> [349,]  68   2      0          7        1   130
#> [350,]  69   8      1          2        0   121
#> [351,]  67  13      1          9        0   130
#> [352,]  68   3      0         19        0   135
#> [353,]  69   1      0          1        1   110
#> [354,]  67   1      0          1        1    60
#> [355,]  67   4      0         60        1   136
#> [356,]  67   8      1          0        1   130
#> [357,]  72  13      1         11        1   195
#> [358,]  68  10      1          8        1   160
#> [359,]  70  35      1          0        1   105
#> [360,]  72  30      1          0        1   145
#> [361,]  70   7      0          7        0   102
#> [362,]  68   7      1          2        0   135
#> [363,]  71   6      0          9        0   120
#> [364,]  70  11      0        180        1   210
#> [365,]  72  19      1          8        0   120
#> [366,]  67   8      0        180        1   170
#> [367,]  67   5      1          0        1   147
#> [368,]  67   9      0        180        0   158
#> [369,]  67   4      1          1        0   134
#> [370,]  72   6      1          5        0   115
#> [371,]  71   1      0        173        1   188
#> [372,]  68  23      0        180        1   220
#> [373,]  70   3      0        180        0   121
#> [374,]  69   3      0        180        0   220
#> [375,]  71   3      1          2        0   150
#> [376,]  68   4      1          3        0   210
#> [377,]  72   5      0         28        0   120
#> [378,]  71   5      0        180        0   191
#> [379,]  69  16      1         10        1   140
#> [380,]  68   7      0        180        1   130
#> [381,]  72  16      1          1        1   130
#> [382,]  70   4      0        180        0   180
#> [383,]  69   1      1          0        0   155
#> [384,]  73   6      1          0        1   270
#> [385,]  72   8      1          1        1   150
#> [386,]  71   2      1          0        1   180
#> [387,]  73   7      0          7        1   140
#> [388,]  68  15      1         13        1   130
#> [389,]  70   3      0          3        1   159
#> [390,]  72   6      0        180        1   130
#> [391,]  73   0      0        180        1   161
#> [392,]  74   8      1          0        1    85
#> [393,]  73   4      0        180        1   154
#> [394,]  71   3      1          1        0   150
#> [395,]  74  20      0         20        1   180
#> [396,]  68   9      0        180        1   120
#> [397,]  71  20      1         10        0   140
#> [398,]  74   0      1          0        1    90
#> [399,]  73   3      1          0        1   136
#> [400,]  71   8      1          7        0   149
#> [401,]  71   3      1          2        1   190
#> [402,]  69  12      1          1        1   149
#> [403,]  74   4      0          4        0   120
#> [404,]  72   5      1          3        1   160
#> [405,]  73   6      0        180        0   110
#> [406,]  72  15      1          0        1   150
#> [407,]  74   3      0          3        1   150
#> [408,]  73  17      1         11        0   140
#> [409,]  71  13      1          8        0   121
#> [410,]  70   4      1          0        1   140
#> [411,]  71  14      1         13        1   170
#> [412,]  74   7      1          0        1   117
#> [413,]  72  10      1          8        1   153
#> [414,]  69   7      0        180        1   144
#> [415,]  72  15      1         13        0   156
#> [416,]  70   8      0          8        0   120
#> [417,]  75   1      0          1        0   133
#> [418,]  73  10      1          9        1   146
#> [419,]  72  10      1          9        1   160
#> [420,]  74  15      1          9        1   179
#> [421,]  71   2      0         10        1   112
#> [422,]  73   1      0          1        1    80
#> [423,]  75   9      1          7        0   140
#> [424,]  75  13      1          1        1   130
#> [425,]  71  11      1          8        0   110
#> [426,]  71   4      0          4        0   134
#> [427,]  72  15      1         12        1   120
#> [428,]  73  10      1          8        0   120
#> [429,]  70   7      1          4        0   184
#> [430,]  72   1      1          1        0   168
#> [431,]  72   7      0         57        1   145
#> [432,]  73  10      0        180        0   162
#> [433,]  72  11      0         11        1   140
#> [434,]  70   3      0          3        0   150
#> [435,]  73   5      1          3        1   112
#> [436,]  76  25      1         12        1   170
#> [437,]  72   2      0        180        0   120
#> [438,]  75   1      0        180        1   140
#> [439,]  72   4      1          0        1   197
#> [440,]  73   5      0        180        0   126
#> [441,]  74  34      1          8        1   233
#> [442,]  76   3      1          0        1   120
#> [443,]  71  32      1         12        1   107
#> [444,]  72   3      0        180        0   160
#> [445,]  76   5      0          5        1   130
#> [446,]  77  11      0         11        1   150
#> [447,]  77   4      0          4        0   185
#> [448,]  75   3      1          1        0   180
#> [449,]  73  15      0         15        1   160
#> [450,]  71  16      0        180        0   140
#> [451,]  73  10      1         10        0   124
#> [452,]  74   3      0          3        1   128
#> [453,]  74   2      1          1        0   140
#> [454,]  76   8      1          0        1   141
#> [455,]  73   6      0          6        1   114
#> [456,]  75  23      1         14        1   110
#> [457,]  72   4      0         85        1   120
#> [458,]  72   4      1          3        0   160
#> [459,]  76  17      1          0        1   200
#> [460,]  73   4      1          3        1   125
#> [461,]  76  13      1         10        0   110
#> [462,]  75   4      1          0        1   122
#> [463,]  73  13      1         11        0   195
#> [464,]  75   4      1          2        1   188
#> [465,]  75   1      0          1        1   125
#> [466,]  74   2      0        180        0   111
#> [467,]  73   1      0         52        1   105
#> [468,]  73   0      0        180        0   156
#> [469,]  78  12      1         11        1   160
#> [470,]  76  44      1         10        0   105
#> [471,]  74  10      1          0        1   135
#> [472,]  74   8      1          8        1   170
#> [473,]  75   9      0        180        1   140
#> [474,]  73  33      1         12        1   175
#> [475,]  77  12      1          9        1   100
#> [476,]  73  10      1          9        0   146
#> [477,]  77  12      0        180        0   130
#> [478,]  77   1      1          0        1    90
#> [479,]  78   5      1          0        1   170
#> [480,]  73   7      1          0        0   174
#> [481,]  74   6      0         79        1   140
#> [482,]  75   3      1          1        1   171
#> [483,]  75   6      0        180        0   150
#> [484,]  79  10      1          8        0   190
#> [485,]  74   2      1          0        1   130
#> [486,]  77   3      0        180        0   110
#> [487,]  74   2      0        180        0   100
#> [488,]  74  15      0        180        1   172
#> [489,]  78   8      1          6        1   110
#> [490,]  74   7      0          7        0   161
#> [491,]  76  13      1          1        1   170
#> [492,]  78  13      1          5        0   130
#> [493,]  75   5      0        119        1   150
#> [494,]  75  12      1          1        1   120
#> [495,]  80   8      0          8        1   120
#> [496,]  75  13      1          6        0   150
#> [497,]  74  10      1          8        0   135
#> [498,]  76   1      0          1        1    83
#> [499,]  79   4      0         80        0   145
#> [500,]  75   4      1          0        0   212
#> [501,]  75   3      0          3        0     0
#> [502,]  76   7      0         29        1   150
#> [503,]  77  24      0         24        1   160
#> [504,]  79   8      0         32        1   120
#> [505,]  80   9      0         23        1   128
#> [506,]  80   6      0          6        1   150
#> [507,]  76   3      1          0        1   140
#> [508,]  79   2      1          0        1   121
#> [509,]  81   1      0          1        0   130
#> [510,]  76   4      0          4        1   160
#> [511,]  79   4      0          4        1   125
#> [512,]  76  10      1          8        0   180
#> [513,]  76  12      1         10        1   127
#> [514,]  80   3      1          0        1   120
#> [515,]  75   2      1          1        1   204
#> [516,]  76   1      0          1        1   140
#> [517,]  77  31      1          3        1   161
#> [518,]  76   1      0          1        1    90
#> [519,]  78   7      1          0        1   110
#> [520,]  79   3      0          3        0   120
#> [521,]  79   4      1          0        1   120
#> [522,]  81   1      0        180        0   120
#> [523,]  80  40      1          0        1   138
#> [524,]  80  17      1         12        0   100
#> [525,]  81   4      1          2        1   126
#> [526,]  79  28      0        164        0   100
#> [527,]  80   9      0        118        1   186
#> [528,]  80   6      0        173        1   160
#> [529,]  79   1      0         37        1   140
#> [530,]  81   3      0        180        0   184
#> [531,]  78   7      0          7        1   147
#> [532,]  77  13      1          0        1   190
#> [533,]  78  15      0         15        0   165
#> [534,]  78   4      0        180        0   175
#> [535,]  81   4      1          1        1   104
#> [536,]  80   1      0          1        0   100
#> [537,]  78   3      1          1        1   152
#> [538,]  77  10      1          8        1   130
#> [539,]  82   3      1          1        1   144
#> [540,]  77   5      0         85        0   188
#> [541,]  80   2      1          1        0   168
#> [542,]  79   6      0          6        0   152
#> [543,]  78   2      0        180        0   148
#> [544,]  82   1      1          0        1    82
#> [545,]  79  10      0        180        1   150
#> [546,]  77   4      0        180        1    98
#> [547,]  78  12      0        180        0   134
#> [548,]  82  21      1          2        0   155
#> [549,]  84  22      1         10        0   180
#> [550,]  79   4      0          4        1   121
#> [551,]  80   6      0          6        1   110
#> [552,]  83   9      1          5        1   170
#> [553,]  82   5      0        180        0   110
#> [554,]  79   7      1          6        0   130
#> [555,]  83   4      0        103        0    97
#> [556,]  81   5      0        177        0    41
#> [557,]  80  11      1          8        0   170
#> [558,]  78   9      1          4        1   120
#> [559,]  79   1      0        180        1   170
#> [560,]  81  15      0        180        1   140
#> [561,]  80   7      1          0        1   146
#> [562,]  83   8      0          8        0   115
#> [563,]  81  16      0         16        1   110
#> [564,]  80   6      1          0        1   150
#> [565,]  81   8      0        180        0   146
#> [566,]  80   8      1          7        0   160
#> [567,]  79   7      0        177        0   197
#> [568,]  79   0      1          0        1    96
#> [569,]  85   4      0        180        0    90
#> [570,]  83   2      0          2        1   155
#> [571,]  82   6      0        128        1   100
#> [572,]  84   4      0        167        0   198
#> [573,]  82  23      1          0        0   110
#> [574,]  84   5      0        180        1   203
#> [575,]  84   4      0          4        1    85
#> [576,]  83   3      0        180        0   174
#> [577,]  79   9      1          8        0   150
#> [578,]  80   2      1          0        1   130
#> [579,]  79   4      0          4        1    60
#> [580,]  80   6      0         71        1   189
#> [581,]  83   1      0          1        1   100
#> [582,]  80  30      1         13        0   220
#> [583,]  79  14      1          0        0   110
#> [584,]  81  14      1         12        1   128
#> [585,]  83   2      0        154        0   130
#> [586,]  85   9      1          6        1   160
#> [587,]  83   1      0        180        0   160
#> [588,]  81   1      0          1        1   145
#> [589,]  81  12      0         12        1   163
#> [590,]  82   5      1          0        1   146
#> [591,]  86  12      0        180        1   120
#> [592,]  83  12      1          2        1   170
#> [593,]  81  19      1         14        0   120
#> [594,]  82   3      1          2        0   130
#> [595,]  80   2      0         88        0   135
#> [596,]  86   3      0          3        1   140
#> [597,]  84   3      0        180        1   120
#> [598,]  85   3      0          3        1   118
#> [599,]  81   2      1          0        1   118
#> [600,]  83   9      0        180        1   149
#> [601,]  83   4      0          4        0   130
#> [602,]  87   2      0          5        1   137
#> [603,]  82  14      1         11        1   103
#> [604,]  86   6      1          0        1   140
#> [605,]  84   3      1          2        0   125
#> [606,]  83  10      1          0        1   190
#> [607,]  86   2      0        180        1   169
#> [608,]  88  14      1          3        1   130
#> [609,]  83  13      1         12        0   170
#> [610,]  84   7      1          2        0   148
#> [611,]  84   9      0         92        1   110
#> [612,]  86   4      0         38        1   122
#> [613,]  82   4      0          4        0   130
#> [614,]  85   3      0          3        1   113
#> [615,]  86   6      0          6        1   117
#> [616,]  84  13      0         62        1   100
#> [617,]  86   6      1          1        0   112
#> [618,]  88   4      0          4        0   100
#> [619,]  83  20      1          3        1   150
#> [620,]  88   4      0          4        1   115
#> [621,]  85  22      0         22        1   184
#> [622,]  83   9      0         65        1   150
#> [623,]  87   2      0        180        1   130
#> [624,]  88   3      0        115        0   110
#> [625,]  88   2      0        180        1    68
#> [626,]  87   8      0          8        1   157
#> [627,]  89   5      0        119        1   140
#> [628,]  87   6      0        180        1   110
#> [629,]  87   1      0          1        0   170
#> [630,]  87  29      0         29        1    97
#> [631,]  87  15      1          9        1   138
#> [632,]  89  10      0         46        1   170
#> [633,]  90  14      0         14        1   100
#> [634,]  88   1      0          1        0   135
#> [635,]  86   4      0        180        1   145
#> [636,]  87   6      1          0        0   125
#> [637,]  91  10      0        145        0   135
#> [638,]  86   3      1          0        1    80
#> [639,]  88   7      0         24        0   119
#> [640,]  87   6      0        126        1   168
#> [641,]  86  10      0        180        1   137
#> [642,]  86   9      1          7        0   130
#> [643,]  87  43      0        178        1   130
#> [644,]  88   3      1          2        0   159
#> [645,]  92   1      0          1        1   167
#> [646,]  91   3      0         33        1   137
#> [647,]  87   7      0         74        1   105
#> [648,]  89  12      1          0        1   130
#> [649,]  91   5      0        169        1   176
#> [650,]  89  52      0         52        1   130
#> [651,]  92   7      0          7        1   110
#> [652,]  89   4      0          4        1   159
#> [653,]  91   0      0          0        0     0
#> [654,]  89  14      0        180        1    84
#> [655,]  90  18      0        180        0   188
#> [656,]  91   4      1          0        1   120
#> [657,]  90  19      1         11        1   129
#> [658,]  94   6      0         50        0    78
#> [659,]  90   1      0          1        1   118
#> [660,]  91   2      0          2        1   116
#> [661,]  94   8      0          8        1   142
#> [662,]  90  16      0         16        1   106
#> [663,]  90   3      0         67        0   162
#> [664,]  96   3      0         12        1    97
#> [665,]  94   3      0         26        1   144
#> [666,]  91  12      0         53        1   212
#> [667,]  93   0      1          0        1   122
#> [668,]  92   2      0          2        0   112
#> [669,]  96   3      1          0        1   104
#> [670,]  96  15      1          0        1   140
#> 
#> $y
#>   [1] 180.0+   5.0+   5.0+ 180.0+ 180.0+   2.0+ 180.0+ 180.0+  12.0    5.0+
#>  [11] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 180.0+ 180.0+ 180.0+   3.0 
#>  [21] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 180.0+ 155.0+ 180.0+
#>  [31] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [41] 180.0+ 180.0+ 150.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [51] 180.0+ 180.0+ 161.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0+ 180.0+
#>  [61] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 172.0+ 180.0+ 180.0+
#>  [71] 180.0+   2.0    1.0  179.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [81] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   7.0   36.0  180.0+
#>  [91]   4.0+ 180.0+ 180.0+  77.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [101]   4.0+ 180.0+ 166.0+  99.0  152.0+   7.0+ 180.0+ 180.0+ 180.0+  13.0+
#> [111] 171.0+ 174.0+  28.0    6.0+ 180.0+ 180.0+ 180.0+ 175.0+   2.0  180.0+
#> [121]   7.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  16.0  180.0+
#> [131] 180.0+ 180.0+  12.0+ 180.0+   2.0  180.0+ 180.0+ 140.0  165.0  180.0+
#> [141] 180.0+ 180.0+ 180.0+   8.0+ 180.0+ 180.0+   0.5  180.0+ 180.0+ 180.0+
#> [151] 171.0+  15.0  166.0+ 180.0+ 147.0+ 180.0+   9.0+   1.0  180.0+ 180.0+
#> [161]   2.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  64.0 
#> [171] 180.0+ 180.0+ 180.0+ 180.0+ 161.0+ 171.0+ 180.0+ 180.0+   3.0    1.0 
#> [181] 180.0+ 180.0+ 180.0+ 180.0+  45.0    3.0+  19.0  180.0+ 180.0+   9.0+
#> [191] 180.0+ 172.0+ 172.0+   8.0    1.0+  15.0  180.0+ 180.0+   8.0+ 180.0+
#> [201] 180.0+ 170.0  180.0+ 169.0    7.0+   6.0  180.0+   3.0+  30.0  180.0+
#> [211]  13.0+   5.0+ 180.0+ 180.0+   1.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [221] 180.0+ 180.0+ 180.0+   9.0    7.0+  22.0  180.0+  84.0    7.0+ 180.0+
#> [231] 180.0+   1.0  180.0+ 180.0+   4.0+   3.0+   6.0+ 180.0+ 180.0+  17.0 
#> [241]  12.0  180.0+  14.0+  36.0  180.0+   3.0+ 180.0+  12.0  180.0+   0.5 
#> [251] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  12.0  180.0+   9.0 
#> [261] 180.0+  11.0+   2.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   3.0+ 103.0 
#> [271] 180.0+   5.0+  13.0  180.0+ 179.0+ 166.0+   3.0    0.5+ 180.0+ 180.0+
#> [281] 175.0+ 180.0+   7.0+   5.0   16.0  180.0+   1.0  180.0+ 123.0+   1.0+
#> [291]  18.0   11.0+ 180.0+  79.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+   8.0+
#> [301] 180.0+  10.0  180.0+ 180.0+ 180.0+ 180.0+  19.0+  12.0   99.0  179.0+
#> [311] 180.0+ 180.0+   0.5   18.0    7.0+ 180.0+ 180.0+ 152.0+ 180.0+ 180.0+
#> [321]  93.0   21.0+ 180.0+   1.0   18.0+   5.0    7.0+ 150.0  180.0+ 180.0+
#> [331] 180.0+ 171.0  180.0+   6.0  180.0+ 180.0+ 180.0+  14.0+ 180.0+   2.0 
#> [341]   5.0+ 103.0    5.0+ 180.0+   7.0  172.0+   7.0  180.0+   7.0    8.0+
#> [351]  13.0+  19.0    1.0    1.0   60.0  180.0+ 132.0   10.0+ 180.0+ 162.0 
#> [361]   7.0+   7.0+   9.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+  76.0  180.0+
#> [371] 173.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  28.0  180.0+  16.0+ 180.0+
#> [381]  16.0+ 180.0+ 180.0+   6.0  180.0+ 180.0+   7.0+  15.0    3.0+ 180.0+
#> [391] 180.0+ 180.0+ 180.0+   3.0+  20.0  180.0+  20.0    0.5  180.0+   8.0 
#> [401]   3.0   12.0    4.0+ 180.0+ 180.0+ 180.0+   3.0  180.0+ 175.0  180.0+
#> [411]  14.0+ 180.0+  10.0+ 180.0+ 180.0+   8.0+   1.0  180.0+ 159.0  180.0+
#> [421]  10.0    1.0  180.0+  13.0  180.0+   4.0+ 180.0+  10.0  104.0+   1.0 
#> [431]  57.0  180.0+  11.0    3.0+   5.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [441]  34.0  180.0+ 177.0+ 180.0+   5.0   11.0+   4.0+ 180.0+  15.0+ 180.0+
#> [451]  10.0    3.0  180.0+ 180.0+   6.0  180.0+  85.0  180.0+  17.0+ 180.0+
#> [461] 174.0+   4.0  180.0+  46.0    1.0  180.0+  52.0  180.0+  12.0  180.0+
#> [471] 180.0+   8.0  180.0+  33.0  180.0+ 180.0+ 180.0+   1.0  180.0+   7.0+
#> [481]  79.0    3.0  180.0+ 180.0+ 176.0+ 180.0+ 180.0+ 180.0+   8.0+   7.0 
#> [491] 180.0+ 172.0  119.0   12.0    8.0  180.0+ 180.0+   1.0   80.0    4.0+
#> [501]   3.0   29.0   24.0   32.0   23.0    6.0    3.0+ 180.0+   1.0    4.0 
#> [511]   4.0   10.0+ 180.0+   3.0+   2.0+   1.0  171.0    1.0   43.0    3.0 
#> [521] 138.0  180.0+  40.0   17.0   93.0  164.0  118.0  173.0   37.0  180.0+
#> [531]   7.0+  22.0   15.0+ 180.0+  71.0    1.0    3.0+  10.0  180.0+  85.0 
#> [541]  10.0    6.0+ 180.0+   1.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0 
#> [551]   6.0    9.0+ 180.0+ 180.0+ 103.0  177.0+ 169.0  180.0+ 180.0+ 180.0+
#> [561]   7.0+   8.0+  16.0  180.0+ 180.0+ 180.0+ 177.0+   0.5  180.0+   2.0 
#> [571] 128.0  167.0   62.0  180.0+   4.0  180.0+ 180.0+ 180.0+   4.0   71.0 
#> [581]   1.0   30.0  180.0+ 180.0+ 154.0  180.0+ 180.0+   1.0   12.0    5.0+
#> [591] 180.0+  77.0  180.0+   3.0   88.0    3.0  180.0+   3.0+ 180.0+ 180.0+
#> [601]   4.0+   5.0  174.0    6.0  180.0+ 180.0+ 180.0+  14.0   13.0  180.0+
#> [611]  92.0   38.0    4.0    3.0+   6.0+  62.0    6.0+   4.0+  20.0    4.0 
#> [621]  22.0   65.0  180.0+ 115.0  180.0+   8.0+ 119.0  180.0+   1.0+  29.0 
#> [631] 180.0+  46.0   14.0    1.0+ 180.0+  25.0  145.0    3.0   24.0  126.0 
#> [641] 180.0+ 180.0+ 178.0+  75.0    1.0   33.0   74.0  180.0+ 169.0   52.0 
#> [651]   7.0    4.0    0.5  180.0+ 180.0+   4.0  180.0+  50.0    1.0+   2.0 
#> [661]   8.0+  16.0   67.0   12.0   26.0   53.0    0.5    2.0    3.0   15.0+
#> 
#> $weights
#> NULL
#> 
#> $offset
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8329575