Cross-Validated GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::cv.glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.cv.glmnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.cv.glmnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.cv_glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

foldid	untyped	NULL		-

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nfolds	integer	10		\([3, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	TRUE	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	lambda.1se		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

standardize.response	logical	FALSE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.gaussian	character	-	covariance, naive	-

type.logistic	character	Newton	Newton, modified.Newton	-

type.measure	character	deviance	deviance, C	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.coxnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvCVGlmnet

Methods

Public methods

LearnerSurvCVGlmnet$new()

LearnerSurvCVGlmnet$selected_features()

LearnerSurvCVGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvCVGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvCVGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvCVGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("surv.cv_glmnet")
print(learner)
#> 
#> ── <LearnerSurvCVGlmnet> (surv.cv_glmnet): Regularized Generalized Linear Model 
#> • Model: -
#> • Parameters: use_pred_offset=TRUE
#> • Packages: mlr3, mlr3proba, mlr3extralearners, and glmnet
#> • Predict Types: [crank], distr, and lp
#> • Feature Types: logical, integer, and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: offset, selected_features, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("grace")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#> Measure: Partial Likelihood Deviance 
#> 
#>      Lambda Index Measure     SE Nonzero
#> min 0.00349    44   2.758 0.1023       6
#> 1se 0.05694    14   2.854 0.1028       4
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  28   9      0        180        1   107
#>   [2,]  32   5      1          0        1   121
#>   [3,]  33   2      0          2        0   150
#>   [4,]  35  10      1          9        0   106
#>   [5,]  35   2      1          1        1   112
#>   [6,]  37   9      0        180        1   151
#>   [7,]  38  13      1          0        1   161
#>   [8,]  38   2      0        115        0   150
#>   [9,]  36   5      1          0        1   115
#>  [10,]  33   6      1          1        1   115
#>  [11,]  38  16      1         10        0   160
#>  [12,]  42   3      1          1        1   130
#>  [13,]  37   1      1          0        1   146
#>  [14,]  40   2      1          1        1   148
#>  [15,]  38   5      1          3        0   125
#>  [16,]  42   2      0          2        0   140
#>  [17,]  40   6      0        180        1   138
#>  [18,]  40  11      1         10        1   120
#>  [19,]  41   2      1          1        0   166
#>  [20,]  40   1      1          0        1   145
#>  [21,]  43   4      1          0        1   130
#>  [22,]  42  15      1         13        1   125
#>  [23,]  40   3      1          1        0   170
#>  [24,]  43   2      1          1        1   116
#>  [25,]  42   2      0        180        1   124
#>  [26,]  44   3      0        180        0   141
#>  [27,]  41  13      1          1        0   140
#>  [28,]  45   6      0        180        1   170
#>  [29,]  41   5      1          4        1   141
#>  [30,]  44   2      1          1        1   150
#>  [31,]  45   2      0        180        1   140
#>  [32,]  46  15      0        180        0   120
#>  [33,]  46   2      1          1        0   126
#>  [34,]  47   4      1          3        0   118
#>  [35,]  45   3      0        150        0   130
#>  [36,]  46   7      1          2        0   166
#>  [37,]  43  29      0        180        1   180
#>  [38,]  45   4      1          0        0   124
#>  [39,]  47   6      1          0        1   116
#>  [40,]  46  13      1         10        0   100
#>  [41,]  47   4      1          3        1   160
#>  [42,]  43   3      1          0        1   124
#>  [43,]  45   8      1          0        1   117
#>  [44,]  49   5      0         73        1   136
#>  [45,]  45   5      0          5        0   141
#>  [46,]  46   6      1          0        1   100
#>  [47,]  44   4      1          0        1   114
#>  [48,]  44   9      1          8        1   135
#>  [49,]  46   4      0        180        1   121
#>  [50,]  44   2      0        180        0   142
#>  [51,]  48   3      0        180        0   154
#>  [52,]  48  12      1         11        0   200
#>  [53,]  47   9      1          6        0   170
#>  [54,]  46   3      1          0        1   119
#>  [55,]  48   2      1          0        0   184
#>  [56,]  47   7      0        180        0   145
#>  [57,]  49   7      1          7        1   110
#>  [58,]  46   3      1          1        1   140
#>  [59,]  46   9      1          9        1   122
#>  [60,]  50   7      0        180        1   110
#>  [61,]  51   1      0          1        1   145
#>  [62,]  47   2      0        180        0   150
#>  [63,]  50   7      1          0        1    92
#>  [64,]  50   4      0          4        1   100
#>  [65,]  50   9      0        180        0   130
#>  [66,]  49   7      1          4        1    90
#>  [67,]  47   8      0        180        0   160
#>  [68,]  52   2      0        180        0   155
#>  [69,]  46   3      0        180        1   120
#>  [70,]  46   1      1          1        0   145
#>  [71,]  50   4      1          1        0   150
#>  [72,]  48   7      1          0        1   110
#>  [73,]  53   8      0         36        1   160
#>  [74,]  48  17      1         10        0   111
#>  [75,]  47   2      1          1        0   110
#>  [76,]  52   4      1          4        0   152
#>  [77,]  49   9      1          3        0   102
#>  [78,]  49  15      0        180        1   160
#>  [79,]  54  17      1         12        1   102
#>  [80,]  53   5      0         77        0   159
#>  [81,]  53   7      1          0        0   199
#>  [82,]  54   6      1          3        0   129
#>  [83,]  51   3      1          1        0   140
#>  [84,]  50   2      0          5        1   106
#>  [85,]  50  10      1          6        0   122
#>  [86,]  50  14      1         13        0   170
#>  [87,]  53   8      1          7        0   160
#>  [88,]  48   3      1          2        0   150
#>  [89,]  51  25      1          1        0   202
#>  [90,]  49   5      1          2        1   150
#>  [91,]  53   4      0          4        0   140
#>  [92,]  52  14      1          7        1   200
#>  [93,]  48   6      0        180        0   160
#>  [94,]  48  11      1         10        0   120
#>  [95,]  54  23      1         10        0   131
#>  [96,]  52   7      1          2        0   154
#>  [97,]  55   6      1          2        1   114
#>  [98,]  54   9      1          1        0   130
#>  [99,]  55   4      1          2        0   150
#> [100,]  50   5      1          4        1   150
#> [101,]  52   4      0        180        0   183
#> [102,]  50   3      0        174        1   153
#> [103,]  49   6      1          0        1   130
#> [104,]  49   1      0          1        1   110
#> [105,]  50   7      1          1        0   156
#> [106,]  53   9      0          9        1    95
#> [107,]  53   8      1          0        1   130
#> [108,]  50   7      1          0        1   127
#> [109,]  52   5      0        175        1   117
#> [110,]  55   1      0        180        0   127
#> [111,]  54   1      0        180        0   162
#> [112,]  56   2      0        180        0   132
#> [113,]  55   5      1          4        1   120
#> [114,]  52   8      0        180        0   119
#> [115,]  53  18      1          9        1   150
#> [116,]  54   3      0        180        1   180
#> [117,]  55   6      0        180        0   170
#> [118,]  52  16      0         16        0   152
#> [119,]  52  16      1         14        0   170
#> [120,]  53  15      0         15        1    90
#> [121,]  53   4      0        180        1   150
#> [122,]  55   6      1          5        1   138
#> [123,]  54  12      1          0        1   190
#> [124,]  55   2      0        134        1   140
#> [125,]  54   3      0        180        0   128
#> [126,]  55   1      0          2        0   130
#> [127,]  57   3      0          3        0   120
#> [128,]  54   7      1          2        0   129
#> [129,]  52   9      1          3        0   170
#> [130,]  54   2      1          1        1   176
#> [131,]  57   5      1          3        1   138
#> [132,]  57   1      0        180        1   156
#> [133,]  57   1      0          1        1   100
#> [134,]  56   4      1          0        1   140
#> [135,]  52   2      0        180        0   140
#> [136,]  52  15      1         14        0   130
#> [137,]  56  14      1         11        0   130
#> [138,]  53   3      1          0        1   200
#> [139,]  57  10      0        180        1   170
#> [140,]  58   8      0          8        1   130
#> [141,]  54   5      0        180        1   108
#> [142,]  55   3      1          1        1   156
#> [143,]  57   0      0          0        1   150
#> [144,]  53  21      1         13        1   130
#> [145,]  53  15      1         10        1   130
#> [146,]  54  17      1          8        1   227
#> [147,]  55  13      0        166        1   140
#> [148,]  56   5      0          5        1   150
#> [149,]  54  23      1          8        0   120
#> [150,]  57   4      1          2        1   185
#> [151,]  53   4      0        147        1   145
#> [152,]  53   7      1          0        1   120
#> [153,]  54   7      1          0        1   141
#> [154,]  59  15      1         10        0   140
#> [155,]  58   1      1          1        1   200
#> [156,]  55   2      0          2        0   106
#> [157,]  59   9      1          1        1   125
#> [158,]  57   1      0        180        0   148
#> [159,]  59   3      0        180        0   120
#> [160,]  57   2      0          2        1   120
#> [161,]  60   5      1          1        0   138
#> [162,]  58  11      1          9        1   124
#> [163,]  55   5      1          0        1   160
#> [164,]  57  10      1          9        0   103
#> [165,]  59   6      1          0        1   140
#> [166,]  59   5      0        180        1   155
#> [167,]  59   4      1          0        1   152
#> [168,]  58   4      1          3        0   120
#> [169,]  60   0      1          0        1    80
#> [170,]  59   2      1          1        0   140
#> [171,]  58   8      0        161        1   140
#> [172,]  61   4      1          3        0   151
#> [173,]  61   9      1          8        0   150
#> [174,]  58   1      0          1        1   100
#> [175,]  57  13      1         10        0   110
#> [176,]  58  10      0         10        1   150
#> [177,]  57   4      1          3        0   138
#> [178,]  61   3      0         17        0   143
#> [179,]  58  19      1         13        1   140
#> [180,]  56  13      1          6        1   158
#> [181,]  56  18      1         11        1   165
#> [182,]  58  11      0        172        1   135
#> [183,]  60  12      1          0        1   114
#> [184,]  56   8      1          8        0   120
#> [185,]  61  13      1         12        1   130
#> [186,]  57   1      0          1        0   126
#> [187,]  59   5      1          2        0   182
#> [188,]  58   5      1          1        1   135
#> [189,]  61  13      0         13        0   210
#> [190,]  58   8      1          5        0   152
#> [191,]  62  10      1          0        1   153
#> [192,]  57   3      1          0        0   100
#> [193,]  61  28      1          7        0   133
#> [194,]  57   7      0        169        0   180
#> [195,]  60   7      0          7        0   147
#> [196,]  59  13      1          2        0   198
#> [197,]  62   4      1          0        0   160
#> [198,]  58   2      0         30        0   202
#> [199,]  59   1      0        180        0   155
#> [200,]  59  16      1          9        1   133
#> [201,]  63   6      0         28        1   120
#> [202,]  61  13      0         13        0   120
#> [203,]  61   5      0          5        1   110
#> [204,]  61   5      0          5        1   160
#> [205,]  58  11      1          9        0   179
#> [206,]  62   1      1          0        1   172
#> [207,]  58   7      0        180        1   150
#> [208,]  63   3      1          1        0   180
#> [209,]  63   1      0        180        1   130
#> [210,]  61   7      0        180        0   135
#> [211,]  62   3      0        180        1   105
#> [212,]  63   4      0        180        1   190
#> [213,]  64   4      0        180        0   130
#> [214,]  63   4      1          1        0   155
#> [215,]  60  18      1         13        0   132
#> [216,]  61   9      1          9        1   150
#> [217,]  62   7      0          7        0   150
#> [218,]  59   1      0         22        1   162
#> [219,]  60   7      1          5        1   141
#> [220,]  60   7      0          7        0   140
#> [221,]  59   5      1          1        0   148
#> [222,]  65  13      0        180        1   100
#> [223,]  63   1      0          1        0   162
#> [224,]  63   1      0          1        0   130
#> [225,]  62   6      0        180        0   170
#> [226,]  61  15      1         13        0   170
#> [227,]  64  10      1          9        0   160
#> [228,]  63  12      1         10        0   200
#> [229,]  59  10      0        180        1   130
#> [230,]  60   8      0         17        1   130
#> [231,]  61   6      1          1        1   117
#> [232,]  64  12      1         11        0   160
#> [233,]  63  10      1          0        1   148
#> [234,]  65  36      1         11        0   140
#> [235,]  66   3      1          1        0   127
#> [236,]  64  32      1          9        1   160
#> [237,]  63  12      1          9        0   114
#> [238,]  63   7      0        180        0   120
#> [239,]  66   5      1          0        1   110
#> [240,]  65   8      1          0        0   168
#> [241,]  65  10      1          8        1   120
#> [242,]  64   0      0          0        1    90
#> [243,]  61  12      1         11        0   154
#> [244,]  64   9      0        180        0   150
#> [245,]  61   4      0        180        1   113
#> [246,]  64   7      0        180        1   120
#> [247,]  66   6      1          1        1   130
#> [248,]  63  12      0         12        1   150
#> [249,]  63   2      1          1        0   180
#> [250,]  62  13      1         11        0   180
#> [251,]  67  11      0         11        1   100
#> [252,]  64   2      0          2        0   201
#> [253,]  66  18      1          5        0   142
#> [254,]  62   9      0        180        0   145
#> [255,]  61  14      1          5        0   140
#> [256,]  61  15      1         10        0   130
#> [257,]  63   3      1          2        0   120
#> [258,]  63   2      1          0        0   140
#> [259,]  64  19      1          8        1   160
#> [260,]  65   8      1          0        1   140
#> [261,]  64  13      1         12        1   150
#> [262,]  64  14      1         13        1   150
#> [263,]  64   0      0          0        1   148
#> [264,]  67   4      1          3        0   130
#> [265,]  66   3      1          0        1   135
#> [266,]  66   6      1          0        1   140
#> [267,]  68   1      0        180        1   166
#> [268,]  64  10      1          9        1   110
#> [269,]  63   7      1          0        0   162
#> [270,]  67   8      1          1        1   130
#> [271,]  68   5      0          5        1    90
#> [272,]  63  10      0         16        1   160
#> [273,]  64   1      0          1        1   120
#> [274,]  68  18      0        180        1   260
#> [275,]  65  17      1         14        1   100
#> [276,]  63   8      1          1        1   162
#> [277,]  63   1      1          0        1   155
#> [278,]  63  10      0         18        1   130
#> [279,]  67  11      0         11        0   150
#> [280,]  68  11      0        180        0   160
#> [281,]  68  14      0         79        0   172
#> [282,]  65  15      1         12        1   150
#> [283,]  66  11      1          0        0   100
#> [284,]  65   4      1          2        1   145
#> [285,]  69  12      0         15        1   140
#> [286,]  66  15      1         13        1   160
#> [287,]  63   2      0        180        0   150
#> [288,]  65  11      1          6        0   130
#> [289,]  69  21      1         10        0   180
#> [290,]  66   9      1          8        0   130
#> [291,]  63   8      0        180        1   120
#> [292,]  68  14      1         13        1   140
#> [293,]  65   8      1          0        1    90
#> [294,]  66   3      0          3        1   138
#> [295,]  69   1      1          0        0   170
#> [296,]  67   1      0        180        1   160
#> [297,]  68  10      1         10        1   150
#> [298,]  67   7      1          4        1   130
#> [299,]  65  10      1          1        1   148
#> [300,]  66  19      1         12        1   150
#> [301,]  67  12      1         12        0   160
#> [302,]  69   6      0         99        1   140
#> [303,]  65   4      1          1        0   130
#> [304,]  64   4      0        179        0   160
#> [305,]  66   4      0        180        1   130
#> [306,]  70  15      1         12        1   132
#> [307,]  64  11      0         11        0   125
#> [308,]  64   0      1          0        1   118
#> [309,]  67   2      0         18        0   131
#> [310,]  66   7      1          5        1   131
#> [311,]  66   4      0        180        0   177
#> [312,]  69   4      1          3        1   150
#> [313,]  65  13      1         12        1   130
#> [314,]  69  17      1         10        0   140
#> [315,]  69   8      0         93        0   140
#> [316,]  65   1      0          1        1   120
#> [317,]  68  18      1          0        1   160
#> [318,]  65   6      0        101        1   115
#> [319,]  68   4      0          4        1   190
#> [320,]  71   3      0          5        0   112
#> [321,]  67   2      0        180        0   128
#> [322,]  66   9      1          3        1   151
#> [323,]  66   1      1          1        1   165
#> [324,]  70   4      1          0        1   180
#> [325,]  69   8      0        180        1   153
#> [326,]  66   4      0        180        0   130
#> [327,]  67  10      1          9        0   200
#> [328,]  67   6      1          4        0   130
#> [329,]  68  18      1         14        1   170
#> [330,]  69   0      0          0        1   148
#> [331,]  68   7      1          0        1   150
#> [332,]  67  14      1         13        0   130
#> [333,]  65  14      1         13        1   150
#> [334,]  69   8      0        180        1   180
#> [335,]  66   2      0          2        1   228
#> [336,]  69   3      0          3        1   130
#> [337,]  70  22      1         13        0   103
#> [338,]  67   5      0          5        0   130
#> [339,]  69   8      1          5        1   195
#> [340,]  69   6      1          4        1   174
#> [341,]  72   7      0          7        1   110
#> [342,]  69   8      1          7        1   108
#> [343,]  67   3      0        180        0   110
#> [344,]  66   2      1          1        0   123
#> [345,]  69  19      0        180        0   130
#> [346,]  67  14      0        172        1   140
#> [347,]  69  11      1          0        1   120
#> [348,]  67   7      1          4        0   122
#> [349,]  69   4      1          3        0   132
#> [350,]  68   2      0          7        1   130
#> [351,]  69   8      1          2        0   121
#> [352,]  67  13      1          9        0   130
#> [353,]  70   3      0        123        0   130
#> [354,]  72   5      1          4        0   170
#> [355,]  67  22      1          1        1   140
#> [356,]  67  12      1          8        0   120
#> [357,]  69   1      0          1        1   110
#> [358,]  67   1      0          1        1    60
#> [359,]  67   8      1          0        1   130
#> [360,]  72  13      1         11        1   195
#> [361,]  70  35      1          0        1   105
#> [362,]  72  30      1          0        1   145
#> [363,]  71   6      0          9        0   120
#> [364,]  69  10      1          6        1   120
#> [365,]  72  19      1          8        0   120
#> [366,]  72  12      1         10        0   170
#> [367,]  67   8      0        180        1   170
#> [368,]  67   5      1          0        1   147
#> [369,]  67   9      0        180        0   158
#> [370,]  73  13      0        152        1   130
#> [371,]  70   5      0        180        0   150
#> [372,]  72   2      0          2        1   100
#> [373,]  67   4      1          1        0   134
#> [374,]  68  23      0        180        1   220
#> [375,]  70   3      0        180        0   121
#> [376,]  69   3      0        180        0   220
#> [377,]  71   3      1          2        0   150
#> [378,]  71   5      0        180        0   191
#> [379,]  73   6      0        180        1   117
#> [380,]  69  16      1         10        1   140
#> [381,]  69   8      1          1        0   164
#> [382,]  68   7      0        180        1   130
#> [383,]  72  16      1          1        1   130
#> [384,]  69   1      1          0        0   155
#> [385,]  73   7      0          7        1   140
#> [386,]  68  15      1         13        1   130
#> [387,]  70  13      1          9        0   100
#> [388,]  74   8      1          0        1    85
#> [389,]  69   2      1          0        1   110
#> [390,]  71   3      1          1        0   150
#> [391,]  71  15      1         11        0   165
#> [392,]  74  20      0         20        1   180
#> [393,]  68   9      0        180        1   120
#> [394,]  71  20      1         10        0   140
#> [395,]  70   5      1          0        1   190
#> [396,]  71  17      1         11        0   160
#> [397,]  71   3      1          2        1   190
#> [398,]  73  10      1          8        0   106
#> [399,]  69  12      1          1        1   149
#> [400,]  70  26      1         11        1   120
#> [401,]  74   4      0          4        0   120
#> [402,]  73   4      0         58        1   160
#> [403,]  70   3      0        180        1   154
#> [404,]  73   6      0        180        0   110
#> [405,]  72  15      1          0        1   150
#> [406,]  72   8      1          0        1   140
#> [407,]  74   3      0          3        1   150
#> [408,]  73  17      1         11        0   140
#> [409,]  71  13      1          8        0   121
#> [410,]  69   2      1          1        1    80
#> [411,]  70   4      1          0        1   140
#> [412,]  71  14      1         13        1   170
#> [413,]  72  10      1          8        1   153
#> [414,]  72  15      1         13        0   156
#> [415,]  71  10      1          9        1   120
#> [416,]  75   2      1          1        0   145
#> [417,]  73  10      1          9        1   146
#> [418,]  74  15      1          9        1   179
#> [419,]  73   1      0          1        1    80
#> [420,]  75   9      1          7        0   140
#> [421,]  75  13      1          1        1   130
#> [422,]  71  11      1          8        0   110
#> [423,]  72  15      1         12        1   120
#> [424,]  73  10      1          8        0   120
#> [425,]  70   7      1          4        0   184
#> [426,]  72   1      1          1        0   168
#> [427,]  73  10      0        180        0   162
#> [428,]  72  11      0         11        1   140
#> [429,]  70   3      0          3        0   150
#> [430,]  73   5      1          3        1   112
#> [431,]  76  25      1         12        1   170
#> [432,]  73  12      1         12        1   140
#> [433,]  72   2      0        180        0   120
#> [434,]  72   4      1          0        1   197
#> [435,]  71   3      1          0        0   144
#> [436,]  73   5      0        180        0   126
#> [437,]  73   4      0        180        0   124
#> [438,]  74  34      1          8        1   233
#> [439,]  72   5      0        180        0   154
#> [440,]  72   3      0        180        0   160
#> [441,]  77  11      0         11        1   150
#> [442,]  77   4      0          4        0   185
#> [443,]  73  15      0         15        1   160
#> [444,]  73  10      1         10        0   124
#> [445,]  76   1      0        180        0   114
#> [446,]  74   2      1          1        0   140
#> [447,]  76   8      1          0        1   141
#> [448,]  74  19      1          4        1   200
#> [449,]  74   2      0        180        0   190
#> [450,]  72   4      0         85        1   120
#> [451,]  72   4      1          3        0   160
#> [452,]  75   0      0          0        1   130
#> [453,]  73  13      1         11        0   195
#> [454,]  75  12      0         12        1   160
#> [455,]  74   8      1          0        1   105
#> [456,]  76  13      1          8        1   148
#> [457,]  75   4      1          2        1   188
#> [458,]  74   6      0        180        0   160
#> [459,]  75   1      0          1        1   125
#> [460,]  74   2      0        180        0   111
#> [461,]  73   1      0         52        1   105
#> [462,]  73   0      0        180        0   156
#> [463,]  78  12      1         11        1   160
#> [464,]  76  44      1         10        0   105
#> [465,]  74  10      1          0        1   135
#> [466,]  75   9      0        180        1   140
#> [467,]  73  33      1         12        1   175
#> [468,]  77  12      1          9        1   100
#> [469,]  73  10      1          9        0   146
#> [470,]  77  12      0        180        0   130
#> [471,]  77   1      1          0        1    90
#> [472,]  76  12      1         11        1   120
#> [473,]  74   6      0         79        1   140
#> [474,]  75   3      1          1        1   171
#> [475,]  75   6      0        180        0   150
#> [476,]  79  10      1          8        0   190
#> [477,]  78  18      0         18        1   144
#> [478,]  77   3      0        180        0   110
#> [479,]  76  29      0         47        0    90
#> [480,]  73   8      1          1        1   162
#> [481,]  73  11      1          2        1   110
#> [482,]  74   2      0        180        0   100
#> [483,]  78   7      0          7        1   133
#> [484,]  74   7      0          7        0   161
#> [485,]  76  13      1          1        1   170
#> [486,]  78  32      1          9        1   198
#> [487,]  78   0      0        180        1   212
#> [488,]  75  12      1          1        1   120
#> [489,]  78  15      0        180        1   270
#> [490,]  74  10      1          8        0   135
#> [491,]  76   1      0          1        1    83
#> [492,]  79   4      0         80        0   145
#> [493,]  78  12      1          9        0   150
#> [494,]  75   4      1          0        0   212
#> [495,]  77   2      1          0        1   143
#> [496,]  78  10      0        180        1   130
#> [497,]  75   3      0          3        0     0
#> [498,]  76   7      0         29        1   150
#> [499,]  77  24      0         24        1   160
#> [500,]  79   8      0         32        1   120
#> [501,]  80   6      0          6        1   150
#> [502,]  78   6      1          0        1   240
#> [503,]  76   3      1          0        1   140
#> [504,]  78  11      1          1        1   140
#> [505,]  79  11      0        180        0   160
#> [506,]  79   2      1          0        1   121
#> [507,]  78  14      1          0        1   140
#> [508,]  79   4      0          4        1   125
#> [509,]  76  10      1          8        0   180
#> [510,]  76  12      1         10        1   127
#> [511,]  77   6      0          6        1   107
#> [512,]  80   3      1          0        1   120
#> [513,]  75   2      1          1        1   204
#> [514,]  78  11      0        180        1   135
#> [515,]  78   7      1          0        1   110
#> [516,]  79   3      0          3        0   120
#> [517,]  77   7      0        180        1   170
#> [518,]  79   4      1          0        1   120
#> [519,]  80  15      1         12        1   150
#> [520,]  82   5      0          8        1   120
#> [521,]  80  40      1          0        1   138
#> [522,]  76   7      0        161        0   151
#> [523,]  80  15      1          0        1    90
#> [524,]  79  28      0        164        0   100
#> [525,]  80   9      0        118        1   186
#> [526,]  80   6      0        173        1   160
#> [527,]  81   3      0        180        0   184
#> [528,]  81   2      0        175        0   172
#> [529,]  77  13      1          0        1   190
#> [530,]  78  15      0         15        0   165
#> [531,]  78   4      0        180        0   175
#> [532,]  78  26      1          5        0   194
#> [533,]  76   1      0        166        0   131
#> [534,]  78  20      1          0        1   109
#> [535,]  80   1      0          1        0   100
#> [536,]  78   3      1          1        1   152
#> [537,]  77  10      1          8        1   130
#> [538,]  82   3      1          1        1   144
#> [539,]  77   5      0         85        0   188
#> [540,]  79   6      0          6        0   152
#> [541,]  80   6      1          0        1   119
#> [542,]  78   2      0        180        0   148
#> [543,]  77   4      0        180        1    98
#> [544,]  81   1      0        108        0   129
#> [545,]  79   1      0        125        0   193
#> [546,]  82  21      1          2        0   155
#> [547,]  79   4      0          4        1   121
#> [548,]  83   9      1          5        1   170
#> [549,]  82   5      0        180        0   110
#> [550,]  83   5      0        180        0   148
#> [551,]  79   7      1          6        0   130
#> [552,]  83   4      0        103        0    97
#> [553,]  81  11      1          8        0   160
#> [554,]  81   5      0        177        0    41
#> [555,]  78   9      1          4        1   120
#> [556,]  82   8      1          1        0   128
#> [557,]  79   1      0        180        1   170
#> [558,]  81  15      0        180        1   140
#> [559,]  80   7      1          0        1   146
#> [560,]  84   5      1          1        1    85
#> [561,]  81  20      1          9        0   170
#> [562,]  83   8      0          8        0   115
#> [563,]  81  16      0         16        1   110
#> [564,]  80   6      1          0        1   150
#> [565,]  79   0      1          0        1    96
#> [566,]  81   2      1          1        0   198
#> [567,]  82   6      0        128        1   100
#> [568,]  84   4      0        167        0   198
#> [569,]  82  23      1          0        0   110
#> [570,]  84   1      0         38        1   205
#> [571,]  79   9      1          8        0   150
#> [572,]  80   2      1          0        1   130
#> [573,]  79   4      0          4        1    60
#> [574,]  80   6      0         71        1   189
#> [575,]  83   1      0          1        1   100
#> [576,]  80  30      1         13        0   220
#> [577,]  83   9      0        180        0   198
#> [578,]  79  14      1          0        0   110
#> [579,]  82   0      0          2        1   100
#> [580,]  85   9      1          6        1   160
#> [581,]  81   4      0          4        0   175
#> [582,]  84  15      1         13        1   110
#> [583,]  81   1      0          1        1   145
#> [584,]  81  12      0         12        1   163
#> [585,]  82  16      1          8        0   103
#> [586,]  81   4      0          4        0   160
#> [587,]  83  12      1          2        1   170
#> [588,]  81  19      1         14        0   120
#> [589,]  82   3      1          2        0   130
#> [590,]  80   2      0         88        0   135
#> [591,]  83   7      0        126        0   135
#> [592,]  86   8      0          8        1   132
#> [593,]  81  16      1          9        0   180
#> [594,]  84   6      0        165        0   145
#> [595,]  86   3      0          3        1   140
#> [596,]  82   9      0        180        1   134
#> [597,]  81  13      0        180        0   152
#> [598,]  85   3      0          3        1   118
#> [599,]  81   4      0        180        0   160
#> [600,]  83   4      0          4        0   130
#> [601,]  87   2      0          5        1   137
#> [602,]  86  12      1          0        1   132
#> [603,]  86   6      1          0        1   140
#> [604,]  84  16      0         70        1   150
#> [605,]  83  19      0         43        0   150
#> [606,]  84   3      1          2        0   125
#> [607,]  86   2      0        180        1   169
#> [608,]  88  14      1          3        1   130
#> [609,]  84   3      0          3        1   121
#> [610,]  83  13      1         12        0   170
#> [611,]  84   7      1          2        0   148
#> [612,]  84   9      0         92        1   110
#> [613,]  84   3      0        180        1   170
#> [614,]  86   4      0         38        1   122
#> [615,]  82   4      0          4        0   130
#> [616,]  84  13      0         62        1   100
#> [617,]  86   6      1          1        0   112
#> [618,]  88   4      0          4        0   100
#> [619,]  83  20      1          3        1   150
#> [620,]  83   9      0         65        1   150
#> [621,]  86   9      1          7        1   142
#> [622,]  87   2      0        180        1   130
#> [623,]  88   3      0        115        0   110
#> [624,]  88   2      0        180        1    68
#> [625,]  83   3      0          3        1   130
#> [626,]  86  15      1          8        1   109
#> [627,]  88   4      0          4        0    86
#> [628,]  89   4      0          4        1   153
#> [629,]  89   5      0        119        1   140
#> [630,]  87   6      0        180        1   110
#> [631,]  84   8      0        180        1   119
#> [632,]  85   8      0          8        1   136
#> [633,]  87  29      0         29        1    97
#> [634,]  84   0      0        180        1   136
#> [635,]  89  10      0         46        1   170
#> [636,]  90  14      0         14        1   100
#> [637,]  87   6      1          0        0   125
#> [638,]  91  10      0        145        0   135
#> [639,]  86   3      1          0        1    80
#> [640,]  88   8      0         50        1   154
#> [641,]  86  10      0        180        1   137
#> [642,]  86   9      1          7        0   130
#> [643,]  90   4      1          0        0   121
#> [644,]  91   1      0          1        1    74
#> [645,]  90   5      1          0        1   125
#> [646,]  88   3      1          2        0   159
#> [647,]  89   3      1          1        1   160
#> [648,]  92   1      0          1        1   167
#> [649,]  91   3      0         33        1   137
#> [650,]  87   7      0         74        1   105
#> [651,]  89  12      1          0        1   130
#> [652,]  89   4      0          4        1   159
#> [653,]  91   0      0          0        0     0
#> [654,]  89  14      0        180        1    84
#> [655,]  90  18      0        180        0   188
#> [656,]  91   4      1          0        1   120
#> [657,]  94   6      0         50        0    78
#> [658,]  90   1      0          1        1   118
#> [659,]  91   2      0          2        1   116
#> [660,]  94   8      0          8        1   142
#> [661,]  92   4      0         76        1   149
#> [662,]  90  16      0         16        1   106
#> [663,]  90   3      0         67        0   162
#> [664,]  96   3      0         12        1    97
#> [665,]  91  12      0         53        1   212
#> [666,]  91   7      0          7        0   135
#> [667,]  93   0      1          0        1   122
#> [668,]  92   2      0          2        0   112
#> [669,]  93   4      0        180        1   135
#> [670,]  96   3      1          0        1   104
#> 
#> $y
#>   [1] 180.0+   5.0+   2.0+ 180.0+   2.0+ 180.0+ 180.0+ 115.0    5.0+ 180.0+
#>  [11] 180.0+ 180.0+ 180.0+   2.0+   5.0+   2.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [21] 180.0+ 180.0+ 180.0+   2.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0+ 180.0+
#>  [31] 180.0+ 180.0+ 180.0+ 180.0+ 150.0  180.0+ 180.0+ 180.0+   6.0+ 180.0+
#>  [41] 180.0+ 180.0+ 180.0+  73.0    5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [51] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   7.0  180.0+ 180.0+ 180.0+
#>  [61]   1.0  180.0+ 180.0+   4.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [71] 180.0+   7.0   36.0   88.0+ 180.0+   4.0+ 180.0+ 180.0+ 180.0+  77.0 
#>  [81] 180.0+ 180.0+ 180.0+   5.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [91]   4.0+  85.0  180.0+ 180.0+ 152.0+   7.0+   6.0+ 180.0+ 180.0+ 171.0+
#> [101] 180.0+ 174.0+   6.0+   1.0  180.0+   9.0+ 180.0+ 180.0+ 175.0+ 180.0+
#> [111] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  16.0+  16.0   15.0+
#> [121] 180.0+ 180.0+  12.0+ 134.0+ 180.0+   2.0    3.0+ 180.0+ 180.0+ 180.0+
#> [131] 140.0  180.0+   1.0  165.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+   8.0+
#> [141] 180.0+ 180.0+   0.5  180.0+ 180.0+ 171.0+ 166.0+   5.0+ 180.0+   4.0+
#> [151] 147.0+ 180.0+ 180.0+ 180.0+   1.0    2.0+ 180.0+ 180.0+ 180.0+   2.0 
#> [161] 180.0+ 180.0+ 180.0+ 180.0+  64.0  180.0+ 180.0+ 180.0+   0.5  180.0+
#> [171] 161.0+ 180.0+ 180.0+   1.0  180.0+  10.0+ 180.0+  17.0   19.0  180.0+
#> [181] 180.0+ 172.0+ 172.0+   8.0  180.0+   1.0+ 180.0+ 180.0+  13.0+   8.0+
#> [191] 180.0+ 180.0+  94.0  169.0    7.0+ 180.0+ 180.0+  30.0  180.0+ 180.0+
#> [201]  28.0   13.0+   5.0    5.0+ 180.0+   1.0  180.0+ 180.0+ 180.0+ 180.0+
#> [211] 180.0+ 180.0+ 180.0+   4.0+ 180.0+ 180.0+   7.0+  22.0   84.0    7.0+
#> [221] 180.0+ 180.0+   1.0    1.0  180.0+ 180.0+ 167.0  180.0+ 180.0+  17.0 
#> [231] 180.0+  12.0  180.0+  36.0    3.0+ 180.0+  12.0  180.0+ 180.0+ 180.0+
#> [241] 180.0+   0.5   12.0+ 180.0+ 180.0+ 180.0+ 180.0+  12.0  180.0+ 180.0+
#> [251]  11.0+   2.0+  18.0+ 180.0+ 180.0+ 180.0+   3.0+   2.0+ 103.0   15.0 
#> [261]  13.0   14.0+   0.5+ 180.0+   3.0+ 180.0+ 180.0+ 180.0+   7.0+   8.0 
#> [271]   5.0   16.0    1.0  180.0+ 180.0+ 180.0+   1.0+  18.0   11.0+ 180.0+
#> [281]  79.0   15.0+ 180.0+   4.0+  15.0  180.0+ 180.0+ 180.0+ 174.0+ 180.0+
#> [291] 180.0+ 180.0+   8.0+   3.0  175.0  180.0+  10.0  180.0+ 180.0+  19.0+
#> [301]  12.0   99.0  180.0+ 179.0+ 180.0+ 180.0+  11.0+   0.5   18.0    7.0+
#> [311] 180.0+ 152.0+ 180.0+ 180.0+  93.0    1.0   18.0+ 101.0    4.0    5.0 
#> [321] 180.0+ 180.0+   1.0  180.0+ 180.0+ 180.0+ 174.0+   6.0  180.0+   0.5 
#> [331] 180.0+ 180.0+  14.0+ 180.0+   2.0    3.0+ 180.0+   5.0+ 180.0+  97.0 
#> [341]   7.0    8.0+ 180.0+   2.0+ 180.0+ 172.0+ 180.0+   7.0  180.0+   7.0 
#> [351]   8.0+  13.0+ 123.0  180.0+  51.0  180.0+   1.0    1.0  180.0+ 132.0 
#> [361] 180.0+ 162.0    9.0  180.0+ 180.0+  12.0  180.0+ 180.0+ 180.0+ 152.0 
#> [371] 180.0+   2.0   76.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  16.0+
#> [381] 180.0+ 180.0+  16.0+ 180.0+   7.0+  15.0   13.0+ 180.0+   2.0    3.0+
#> [391] 180.0+  20.0  180.0+  20.0  180.0+ 180.0+   3.0   87.0   12.0  180.0+
#> [401]   4.0+  58.0  180.0+ 180.0+ 180.0+ 180.0+   3.0  180.0+ 175.0    2.0 
#> [411] 180.0+  14.0+  10.0+ 180.0+ 179.0+ 180.0+ 180.0+ 180.0+   1.0  180.0+
#> [421]  13.0  180.0+ 180.0+  10.0  104.0+   1.0  180.0+  11.0    3.0+   5.0 
#> [431] 180.0+  12.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+  34.0  180.0+ 180.0+
#> [441]  11.0+   4.0+  15.0+  10.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+  85.0 
#> [451] 180.0+   0.5  180.0+  12.0  180.0+ 180.0+  46.0  180.0+   1.0  180.0+
#> [461]  52.0  180.0+  12.0  180.0+ 180.0+ 180.0+  33.0  180.0+ 180.0+ 180.0+
#> [471]   1.0   12.0   79.0    3.0  180.0+ 180.0+  18.0  180.0+  47.0  180.0+
#> [481]  11.0  180.0+   7.0    7.0  180.0+  32.0  180.0+  12.0  180.0+ 180.0+
#> [491]   1.0   80.0  180.0+   4.0+   2.0  180.0+   3.0   29.0   24.0   32.0 
#> [501]   6.0  180.0+   3.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0   10.0+ 180.0+
#> [511]   6.0    3.0+   2.0+ 180.0+  43.0    3.0  180.0+ 138.0  180.0+   8.0 
#> [521]  40.0  161.0  180.0+ 164.0  118.0  173.0  180.0+ 175.0+  22.0   15.0+
#> [531] 180.0+ 171.0+ 166.0+  20.0+   1.0    3.0+  10.0  180.0+  85.0    6.0+
#> [541]   6.0  180.0+ 180.0+ 108.0  125.0  180.0+   4.0    9.0+ 180.0+ 180.0+
#> [551] 180.0+ 103.0  180.0+ 177.0+ 180.0+ 180.0+ 180.0+ 180.0+   7.0+ 180.0+
#> [561]  20.0    8.0+  16.0  180.0+   0.5  180.0+ 128.0  167.0   62.0   38.0 
#> [571] 180.0+ 180.0+   4.0   71.0    1.0   30.0  180.0+ 180.0+   2.0  180.0+
#> [581]   4.0+ 180.0+   1.0   12.0   16.0+   4.0+  77.0  180.0+   3.0   88.0 
#> [591] 126.0    8.0  180.0+ 165.0    3.0  180.0+ 180.0+   3.0+ 180.0+   4.0+
#> [601]   5.0  180.0+   6.0   70.0   43.0  180.0+ 180.0+  14.0    3.0   13.0 
#> [611] 180.0+  92.0  180.0+  38.0    4.0   62.0    6.0+   4.0+  20.0   65.0 
#> [621]  11.0  180.0+ 115.0  180.0+   3.0+ 180.0+   4.0    4.0  119.0  180.0+
#> [631] 180.0+   8.0   29.0  180.0+  46.0   14.0   25.0  145.0    3.0   50.0 
#> [641] 180.0+ 180.0+   4.0    1.0   89.0   75.0    3.0+   1.0   33.0   74.0 
#> [651] 180.0+   4.0    0.5  180.0+ 180.0+   4.0   50.0    1.0+   2.0    8.0+
#> [661]  76.0   16.0   67.0   12.0   53.0    7.0+   0.5    2.0  180.0+   3.0 
#> 
#> $weights
#> NULL
#> 
#> $offset
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8425726