Cross-Validated GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::cv.glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.cv.glmnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.cv.glmnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.cv_glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

foldid	untyped	NULL		-

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nfolds	integer	10		\([3, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	TRUE	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	lambda.1se		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

standardize.response	logical	FALSE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.gaussian	character	-	covariance, naive	-

type.logistic	character	Newton	Newton, modified.Newton	-

type.measure	character	deviance	deviance, C	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.coxnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvCVGlmnet

Methods

Public methods

LearnerSurvCVGlmnet$new()

LearnerSurvCVGlmnet$selected_features()

LearnerSurvCVGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvCVGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvCVGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvCVGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("surv.cv_glmnet")
print(learner)
#> 
#> ── <LearnerSurvCVGlmnet> (surv.cv_glmnet): Regularized Generalized Linear Model 
#> • Model: -
#> • Parameters: use_pred_offset=TRUE
#> • Packages: mlr3, mlr3proba, mlr3extralearners, and glmnet
#> • Predict Types: [crank], distr, and lp
#> • Feature Types: logical, integer, and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: offset, selected_features, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("grace")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#> Measure: Partial Likelihood Deviance 
#> 
#>      Lambda Index Measure     SE Nonzero
#> min 0.00349    45   2.779 0.1496       6
#> 1se 0.06853    13   2.918 0.1693       5
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  35   5      1          2        0   172
#>   [2,]  35  10      1          9        0   106
#>   [3,]  34   5      0          5        0   120
#>   [4,]  35   2      0        180        0   121
#>   [5,]  35   2      1          1        1   112
#>   [6,]  37   9      0        180        1   151
#>   [7,]  38  13      1          0        1   161
#>   [8,]  38   2      0        115        0   150
#>   [9,]  35   0      0        180        1   119
#>  [10,]  38  12      1          8        1   120
#>  [11,]  33   6      1          1        1   115
#>  [12,]  38  12      1         11        1    92
#>  [13,]  37   1      1          0        1   146
#>  [14,]  40   2      1          1        1   148
#>  [15,]  42   2      0        180        1   100
#>  [16,]  38   5      1          3        0   125
#>  [17,]  42   2      0          2        0   140
#>  [18,]  40   6      0        180        1   138
#>  [19,]  40  11      1         10        1   120
#>  [20,]  42   2      0        180        0   100
#>  [21,]  43   3      1          0        1   100
#>  [22,]  41   2      1          1        0   166
#>  [23,]  42   4      0        180        0   162
#>  [24,]  42  15      1         13        1   125
#>  [25,]  43   2      1          1        1   116
#>  [26,]  42   2      0        180        1   124
#>  [27,]  44   5      1          1        0   170
#>  [28,]  45   3      0        180        1   154
#>  [29,]  41  10      1          8        0   150
#>  [30,]  41  13      1          1        0   140
#>  [31,]  45   2      0        180        1   140
#>  [32,]  46  15      0        180        0   120
#>  [33,]  47   4      1          3        0   118
#>  [34,]  44   3      1          0        1   180
#>  [35,]  46   7      1          2        0   166
#>  [36,]  43  29      0        180        1   180
#>  [37,]  45   4      1          0        0   124
#>  [38,]  47   6      1          0        1   116
#>  [39,]  44   0      1          0        1    96
#>  [40,]  43   3      1          0        1   124
#>  [41,]  45   8      1          0        1   117
#>  [42,]  45   5      0          5        0   141
#>  [43,]  46   6      1          0        1   100
#>  [44,]  44   9      1          8        1   135
#>  [45,]  46   4      0        180        1   121
#>  [46,]  46  15      0        180        1   120
#>  [47,]  45   9      1          0        1   145
#>  [48,]  48   3      0        180        0   154
#>  [49,]  47   5      1          3        1   130
#>  [50,]  47   9      1          6        0   170
#>  [51,]  46   3      1          0        1   119
#>  [52,]  49   4      0        180        0   117
#>  [53,]  48   2      1          0        0   184
#>  [54,]  50   6      1          2        1   140
#>  [55,]  49   7      1          7        1   110
#>  [56,]  46   3      1          1        1   140
#>  [57,]  46   9      1          9        1   122
#>  [58,]  50   7      0        180        1   110
#>  [59,]  49  15      1         11        1   160
#>  [60,]  47   2      0        180        0   150
#>  [61,]  49  23      0        179        1   112
#>  [62,]  46   6      1          0        1   156
#>  [63,]  52   2      0        180        1   170
#>  [64,]  50   7      1          0        1    92
#>  [65,]  51   3      1          2        0   113
#>  [66,]  49   7      1          4        1    90
#>  [67,]  47   8      0        180        0   160
#>  [68,]  47   6      0        180        1   162
#>  [69,]  51   8      0        180        1   140
#>  [70,]  46   3      0        180        1   120
#>  [71,]  46   1      1          1        0   145
#>  [72,]  53   8      0         36        1   160
#>  [73,]  47   2      1          1        0   110
#>  [74,]  49  15      0        180        1   160
#>  [75,]  53   5      0        180        1   140
#>  [76,]  54  17      1         12        1   102
#>  [77,]  53   5      0         77        0   159
#>  [78,]  53   7      1          0        0   199
#>  [79,]  51   3      1          1        0   140
#>  [80,]  50   2      0          5        1   106
#>  [81,]  50  10      1          6        0   122
#>  [82,]  50  14      1         13        0   170
#>  [83,]  48   3      1          2        0   150
#>  [84,]  51  25      1          1        0   202
#>  [85,]  49   5      1          2        1   150
#>  [86,]  52  14      1          7        1   200
#>  [87,]  48   6      0        180        0   160
#>  [88,]  48  11      1         10        0   120
#>  [89,]  53   4      1          0        1   156
#>  [90,]  54   9      1          0        1   138
#>  [91,]  49  16      0         16        0   125
#>  [92,]  54  23      1         10        0   131
#>  [93,]  52   7      1          2        0   154
#>  [94,]  55   6      1          2        1   114
#>  [95,]  55   4      1          2        0   150
#>  [96,]  52   4      0        180        1   180
#>  [97,]  51  13      1         11        0   145
#>  [98,]  50   5      1          4        1   150
#>  [99,]  52   4      0        180        0   183
#> [100,]  50   3      0        174        1   153
#> [101,]  55  28      1         13        1   160
#> [102,]  49   6      1          0        1   130
#> [103,]  49   1      0          1        1   110
#> [104,]  53   9      0          9        1    95
#> [105,]  50   7      1          0        1   127
#> [106,]  56   4      1          1        1   130
#> [107,]  54   1      0        180        0   162
#> [108,]  54   7      1          0        1   100
#> [109,]  56   3      0        180        1   193
#> [110,]  53  18      1          9        1   150
#> [111,]  54   3      0        180        1   180
#> [112,]  55   6      0        180        0   170
#> [113,]  52  16      0         16        0   152
#> [114,]  52  16      1         14        0   170
#> [115,]  53  15      0         15        1    90
#> [116,]  53   4      0        180        1   150
#> [117,]  54  12      1          0        1   190
#> [118,]  55   2      0        134        1   140
#> [119,]  56   3      0          8        1   139
#> [120,]  55   1      0          2        0   130
#> [121,]  57   3      0          3        0   120
#> [122,]  54   7      1          2        0   129
#> [123,]  54   2      1          1        0   135
#> [124,]  57   1      0        180        1   156
#> [125,]  55  11      1          7        0   104
#> [126,]  52  15      1         14        0   130
#> [127,]  56  14      1         11        0   130
#> [128,]  57  10      0        180        1   170
#> [129,]  54   5      0        180        1   108
#> [130,]  55   3      1          1        1   156
#> [131,]  57   0      0          0        1   150
#> [132,]  53  21      1         13        1   130
#> [133,]  57   4      0        180        1   119
#> [134,]  58   6      1          0        1    90
#> [135,]  54  17      1          8        1   227
#> [136,]  55   9      1          2        1   147
#> [137,]  55  13      0        166        1   140
#> [138,]  56   5      0          5        1   150
#> [139,]  54  23      1          8        0   120
#> [140,]  57   4      1          2        1   185
#> [141,]  53   4      0        147        1   145
#> [142,]  53   7      1          0        1   120
#> [143,]  57  11      1         10        1   129
#> [144,]  55   3      1          2        0   140
#> [145,]  55   5      0          5        1   131
#> [146,]  54   7      1          0        1   141
#> [147,]  56   4      0          4        0   164
#> [148,]  59  15      1         10        0   140
#> [149,]  58   9      1          0        1   180
#> [150,]  58   1      1          1        1   200
#> [151,]  55   5      1          0        0   140
#> [152,]  56   7      1          5        1   120
#> [153,]  55   2      0          2        0   106
#> [154,]  59   9      1          1        1   125
#> [155,]  60  11      1          9        0   106
#> [156,]  59   3      0        180        0   120
#> [157,]  58   4      1          0        1   160
#> [158,]  60   5      1          1        0   138
#> [159,]  57   5      0        180        1   130
#> [160,]  58  11      1          9        1   124
#> [161,]  55   5      1          0        1   160
#> [162,]  57  10      1          9        0   103
#> [163,]  59   6      1          0        1   140
#> [164,]  59   5      0        180        1   155
#> [165,]  59   4      1          0        1   152
#> [166,]  58   4      1          3        0   120
#> [167,]  59   2      1          1        0   140
#> [168,]  58  14      1          6        0   190
#> [169,]  58   1      0          1        1   100
#> [170,]  61  20      1         13        0   130
#> [171,]  57  13      1         10        0   110
#> [172,]  57   2      1          0        1   116
#> [173,]  58  10      0         10        1   150
#> [174,]  57   4      1          3        0   138
#> [175,]  61   3      0         17        0   143
#> [176,]  56  14      0         45        0   130
#> [177,]  57   3      1          2        0   120
#> [178,]  58  19      1         13        1   140
#> [179,]  56  18      1         11        1   165
#> [180,]  55   4      1          3        1   160
#> [181,]  58  11      0        172        1   135
#> [182,]  60  12      1          0        1   114
#> [183,]  55   9      1          7        1   135
#> [184,]  61   4      1          0        1   115
#> [185,]  59  11      1          8        1   190
#> [186,]  57   1      0          1        0   126
#> [187,]  59   5      1          2        0   182
#> [188,]  58   5      1          1        1   135
#> [189,]  61   8      0         77        0   120
#> [190,]  61  13      0         13        0   210
#> [191,]  58   8      1          5        0   152
#> [192,]  62  10      1          0        1   153
#> [193,]  62   7      1          2        1   180
#> [194,]  61  28      1          7        0   133
#> [195,]  58   8      1          3        1   150
#> [196,]  57   7      0        169        0   180
#> [197,]  61   7      0          7        1   150
#> [198,]  60   7      0          7        0   147
#> [199,]  61   6      0          6        0   134
#> [200,]  59  13      1          2        0   198
#> [201,]  62   4      1          0        0   160
#> [202,]  60  17      1          8        1   140
#> [203,]  58   3      1          0        1   146
#> [204,]  62   4      1          3        0   173
#> [205,]  58   2      0         30        0   202
#> [206,]  59   1      0        180        0   155
#> [207,]  59  16      1          9        1   133
#> [208,]  61  13      0         13        0   120
#> [209,]  61   5      0          5        1   110
#> [210,]  57  18      1          9        1    93
#> [211,]  61   5      0          5        1   160
#> [212,]  58  11      1          9        0   179
#> [213,]  57   2      1          1        0   159
#> [214,]  62   1      1          0        1   172
#> [215,]  63   3      1          1        0   180
#> [216,]  61   7      0        180        0   135
#> [217,]  63   4      1          3        0   222
#> [218,]  63   4      1          1        0   155
#> [219,]  59   8      0        180        1   140
#> [220,]  60   7      1          5        1   141
#> [221,]  60   7      0          7        0   140
#> [222,]  60   7      1          1        1    90
#> [223,]  65  13      0        180        1   100
#> [224,]  63   1      0          1        0   162
#> [225,]  62   6      0        180        0   170
#> [226,]  60   3      0          3        0   168
#> [227,]  62   6      0          6        0   120
#> [228,]  63  12      1         10        0   200
#> [229,]  61   6      1          1        1   117
#> [230,]  66   1      1          0        1   120
#> [231,]  63  14      1          9        0   123
#> [232,]  66   3      1          1        0   127
#> [233,]  61  10      1          2        1   194
#> [234,]  63  12      1          9        0   114
#> [235,]  63   7      0        180        0   120
#> [236,]  66   5      1          0        1   110
#> [237,]  65   8      1          0        0   168
#> [238,]  65  10      1          8        1   120
#> [239,]  64   0      0          0        1    90
#> [240,]  60   6      0        180        0   130
#> [241,]  64  21      1         10        0   190
#> [242,]  61  12      1         11        0   154
#> [243,]  61   4      0        180        1   113
#> [244,]  65   3      0        180        1   190
#> [245,]  64   7      0        180        1   120
#> [246,]  66   6      1          1        1   130
#> [247,]  63  12      0         12        1   150
#> [248,]  62   3      1          1        1   199
#> [249,]  65   6      0          9        0   112
#> [250,]  65   3      1          0        1    80
#> [251,]  63   5      1          4        0   170
#> [252,]  62  13      1         11        0   180
#> [253,]  66  18      1          5        0   142
#> [254,]  66  16      1         11        1   169
#> [255,]  62   9      0        180        0   145
#> [256,]  61  14      1          5        0   140
#> [257,]  61  15      1         10        0   130
#> [258,]  63   9      1          8        1   160
#> [259,]  63   3      1          2        0   120
#> [260,]  63   2      1          0        0   140
#> [261,]  65   8      1          0        1   140
#> [262,]  67   6      0        180        1   170
#> [263,]  65  15      1         11        1   160
#> [264,]  64  13      1         12        1   150
#> [265,]  66   7      1          0        1   115
#> [266,]  66  13      1          0        0   118
#> [267,]  64  14      1         13        1   150
#> [268,]  65   3      0          3        0   105
#> [269,]  64   0      0          0        1   148
#> [270,]  67   4      1          3        0   130
#> [271,]  66   6      1          0        1   140
#> [272,]  65   2      1          1        1   170
#> [273,]  68   1      0        180        1   166
#> [274,]  63   7      1          0        0   162
#> [275,]  67   8      1          1        1   130
#> [276,]  68   5      0          5        1    90
#> [277,]  63  10      0         16        1   160
#> [278,]  66  14      0        180        0   130
#> [279,]  64   1      0          1        1   120
#> [280,]  68  18      0        180        1   260
#> [281,]  65  17      1         14        1   100
#> [282,]  65  18      1          3        0   120
#> [283,]  63   1      1          0        1   155
#> [284,]  67  11      0         11        0   150
#> [285,]  68  14      0         79        0   172
#> [286,]  66  12      1         10        1   150
#> [287,]  65  15      1         12        1   150
#> [288,]  65   4      1          2        1   145
#> [289,]  66  15      1         13        1   160
#> [290,]  63   2      0        180        0   150
#> [291,]  65  11      1          6        0   130
#> [292,]  69   6      0        180        1   100
#> [293,]  63   8      0        180        1   120
#> [294,]  68  14      1         13        1   140
#> [295,]  65   8      1          0        1    90
#> [296,]  66   3      0          3        1   138
#> [297,]  68  10      1         10        1   150
#> [298,]  65   1      1          0        0   133
#> [299,]  65  10      1          1        1   148
#> [300,]  66  19      1         12        1   150
#> [301,]  67  12      1         12        0   160
#> [302,]  69   6      0         99        1   140
#> [303,]  65   4      1          1        0   130
#> [304,]  64   4      0        179        0   160
#> [305,]  66   4      0        180        1   130
#> [306,]  64   0      1          0        1   118
#> [307,]  67   2      0         18        0   131
#> [308,]  66   4      0        180        0   177
#> [309,]  68   4      1          0        1   160
#> [310,]  69   4      1          3        1   150
#> [311,]  69  17      1         10        0   140
#> [312,]  66   6      0        180        0   140
#> [313,]  65   1      0          1        1   120
#> [314,]  68  18      1          0        1   160
#> [315,]  65   6      0        101        1   115
#> [316,]  68   4      0          4        1   190
#> [317,]  70   7      1          0        1   190
#> [318,]  68   7      0        150        0   210
#> [319,]  71  20      1          0        1   160
#> [320,]  66   9      1          3        1   151
#> [321,]  70   4      1          0        1   180
#> [322,]  69   8      0        180        1   153
#> [323,]  70  14      0        171        0   166
#> [324,]  66   4      0        180        0   130
#> [325,]  67   6      1          4        0   130
#> [326,]  68  18      1         14        1   170
#> [327,]  69   0      0          0        1   148
#> [328,]  65   2      0        180        0   130
#> [329,]  68   7      1          0        1   150
#> [330,]  69   3      1          2        0   151
#> [331,]  65  14      1         13        1   150
#> [332,]  69   8      0        180        1   180
#> [333,]  71   7      0          7        0   230
#> [334,]  66   2      0          2        1   228
#> [335,]  71   6      0         45        1   158
#> [336,]  69   5      0          5        1   142
#> [337,]  68   6      0        180        0   145
#> [338,]  69   8      1          5        1   195
#> [339,]  72   3      1          0        1   132
#> [340,]  72   7      0          7        1   110
#> [341,]  67   3      0        180        0   110
#> [342,]  69  19      0        180        0   130
#> [343,]  68  18      0         18        1   100
#> [344,]  67  14      0        172        1   140
#> [345,]  69  11      1          0        1   120
#> [346,]  66   2      0        180        0   130
#> [347,]  67   7      1          4        0   122
#> [348,]  69   4      1          3        0   132
#> [349,]  69   8      1          2        0   121
#> [350,]  67  13      1          9        0   130
#> [351,]  70   3      0        123        0   130
#> [352,]  70   9      0        180        1   142
#> [353,]  72   5      1          4        0   170
#> [354,]  68   3      0         19        0   135
#> [355,]  67  12      1          8        0   120
#> [356,]  69   1      0          1        1   110
#> [357,]  67   1      0          1        1    60
#> [358,]  67   4      0         60        1   136
#> [359,]  69   5      0         76        0   120
#> [360,]  66  24      1         13        0   130
#> [361,]  70  35      1          0        1   105
#> [362,]  70   7      0          7        0   102
#> [363,]  68   7      1          2        0   135
#> [364,]  73  20      1          0        1   170
#> [365,]  70  11      0        180        1   210
#> [366,]  72  19      1          8        0   120
#> [367,]  70   5      0        180        0   150
#> [368,]  72   6      1          5        0   115
#> [369,]  68  23      0        180        1   220
#> [370,]  70   3      0        180        0   121
#> [371,]  71   3      1          2        0   150
#> [372,]  68   4      1          3        0   210
#> [373,]  72   5      0         28        0   120
#> [374,]  71   5      0        180        0   191
#> [375,]  73   6      0        180        1   117
#> [376,]  69  16      1         10        1   140
#> [377,]  69   8      1          1        0   164
#> [378,]  68   7      0        180        1   130
#> [379,]  72  16      1          1        1   130
#> [380,]  69   1      1          0        0   155
#> [381,]  73   6      1          0        1   270
#> [382,]  72   8      1          1        1   150
#> [383,]  71   2      1          0        1   180
#> [384,]  68  15      1         13        1   130
#> [385,]  72   6      0        180        1   130
#> [386,]  74   8      1          0        1    85
#> [387,]  73   4      0        180        1   154
#> [388,]  71   3      1          1        0   150
#> [389,]  71  15      1         11        0   165
#> [390,]  74  20      0         20        1   180
#> [391,]  68   9      0        180        1   120
#> [392,]  74   0      1          0        1    90
#> [393,]  73   3      1          0        1   136
#> [394,]  71   8      1          7        0   149
#> [395,]  71   3      1          2        1   190
#> [396,]  73  10      1          8        0   106
#> [397,]  74   4      0          4        0   120
#> [398,]  73   6      0        180        0   110
#> [399,]  72  15      1          0        1   150
#> [400,]  74   3      0          3        1   150
#> [401,]  73  17      1         11        0   140
#> [402,]  71  13      1          8        0   121
#> [403,]  69   2      1          1        1    80
#> [404,]  71  14      1         13        1   170
#> [405,]  69   7      0        180        1   144
#> [406,]  72  15      1         13        0   156
#> [407,]  75   1      0          1        0   133
#> [408,]  73  10      1          9        1   146
#> [409,]  72  10      1          9        1   160
#> [410,]  74  15      1          9        1   179
#> [411,]  71   2      0         10        1   112
#> [412,]  73   1      0          1        1    80
#> [413,]  75  13      1          1        1   130
#> [414,]  71   4      0          4        0   134
#> [415,]  72  15      1         12        1   120
#> [416,]  73  10      1          8        0   120
#> [417,]  72   1      1          1        0   168
#> [418,]  72   7      0         57        1   145
#> [419,]  73  10      0        180        0   162
#> [420,]  72  11      0         11        1   140
#> [421,]  76  25      1         12        1   170
#> [422,]  75   1      0        180        1   140
#> [423,]  71   3      1          0        0   144
#> [424,]  73   5      0        180        0   126
#> [425,]  76   3      1          0        1   120
#> [426,]  71  32      1         12        1   107
#> [427,]  72   5      0        180        0   154
#> [428,]  77   4      0          4        0   185
#> [429,]  75   3      1          1        0   180
#> [430,]  71  16      0        180        0   140
#> [431,]  74   7      0        180        1   150
#> [432,]  74   3      0          3        1   128
#> [433,]  76   1      0        180        0   114
#> [434,]  74   2      1          1        0   140
#> [435,]  76   8      1          0        1   141
#> [436,]  74  19      1          4        1   200
#> [437,]  73   6      0          6        1   114
#> [438,]  75  23      1         14        1   110
#> [439,]  74   2      0        180        0   190
#> [440,]  72   4      0         85        1   120
#> [441,]  72   4      1          3        0   160
#> [442,]  76  17      1          0        1   200
#> [443,]  73   4      1          3        1   125
#> [444,]  75   7      0          7        0   190
#> [445,]  75   0      0          0        1   130
#> [446,]  73  13      1         11        0   195
#> [447,]  75  12      0         12        1   160
#> [448,]  74   8      1          0        1   105
#> [449,]  75   4      1          2        1   188
#> [450,]  76   4      0          4        1   155
#> [451,]  75   1      0          1        1   125
#> [452,]  74   2      0        180        0   111
#> [453,]  73   1      0         52        1   105
#> [454,]  72   5      0        180        0   120
#> [455,]  78  12      1         11        1   160
#> [456,]  76   5      0        180        0   185
#> [457,]  74  10      1          0        1   135
#> [458,]  76   5      1          0        1   167
#> [459,]  74   8      1          8        1   170
#> [460,]  75   9      0        180        1   140
#> [461,]  77  12      1          9        1   100
#> [462,]  77   1      1          0        1    90
#> [463,]  76  12      1         11        1   120
#> [464,]  78   5      1          0        1   170
#> [465,]  73   7      1          0        0   174
#> [466,]  74   6      0         79        1   140
#> [467,]  74   9      1          8        0   200
#> [468,]  75   6      0        180        0   150
#> [469,]  79  10      1          8        0   190
#> [470,]  74   2      1          0        1   130
#> [471,]  76  29      0         47        0    90
#> [472,]  73   8      1          1        1   162
#> [473,]  73  11      1          2        1   110
#> [474,]  78   7      0          7        1   133
#> [475,]  78   8      1          6        1   110
#> [476,]  74   7      0          7        0   161
#> [477,]  78  32      1          9        1   198
#> [478,]  79   6      0        180        0   170
#> [479,]  78   0      0        180        1   212
#> [480,]  78  13      1          5        0   130
#> [481,]  75   5      0        119        1   150
#> [482,]  75  12      1          1        1   120
#> [483,]  78  15      0        180        1   270
#> [484,]  80   8      0          8        1   120
#> [485,]  76   1      0          1        1    83
#> [486,]  78  12      1          9        0   150
#> [487,]  78   2      1          1        0   130
#> [488,]  75   4      1          0        0   212
#> [489,]  77   2      1          0        1   143
#> [490,]  76  11      1          0        0   120
#> [491,]  75  11      1          4        0   162
#> [492,]  75   3      0          3        0     0
#> [493,]  77  24      0         24        1   160
#> [494,]  79   8      0         32        1   120
#> [495,]  80   6      0          6        1   150
#> [496,]  79  11      0        180        0   160
#> [497,]  78  14      1          0        1   140
#> [498,]  81   1      0          1        0   130
#> [499,]  78  11      1          8        1   118
#> [500,]  76   4      0          4        1   160
#> [501,]  79   4      0          4        1   125
#> [502,]  76  10      1          8        0   180
#> [503,]  77   6      0          6        1   107
#> [504,]  78  11      0        180        1   135
#> [505,]  76   1      0          1        1   140
#> [506,]  77  31      1          3        1   161
#> [507,]  76   1      0          1        1    90
#> [508,]  78   7      1          0        1   110
#> [509,]  79   3      0          3        0   120
#> [510,]  77   6      0          6        1   144
#> [511,]  79   4      1          0        1   120
#> [512,]  77   9      1          4        0   141
#> [513,]  82   5      0          8        1   120
#> [514,]  80  40      1          0        1   138
#> [515,]  78   4      0         59        1   112
#> [516,]  80  17      1         12        0   100
#> [517,]  76   7      0        161        0   151
#> [518,]  79  10      0         10        1   120
#> [519,]  81   4      1          2        1   126
#> [520,]  80   9      0        118        1   186
#> [521,]  80   6      0        173        1   160
#> [522,]  81   3      0        180        0   184
#> [523,]  78   7      0          7        1   147
#> [524,]  77  13      1          0        1   190
#> [525,]  78  15      0         15        0   165
#> [526,]  79   3      0          3        1   101
#> [527,]  78  26      1          5        0   194
#> [528,]  76   1      0        166        0   131
#> [529,]  80   1      0          1        0   100
#> [530,]  78   3      1          1        1   152
#> [531,]  77  10      1          8        1   130
#> [532,]  82   3      1          1        1   144
#> [533,]  77   5      0         85        0   188
#> [534,]  80   2      1          1        0   168
#> [535,]  79   6      0          6        0   152
#> [536,]  80   6      1          0        1   119
#> [537,]  78   2      0        180        0   148
#> [538,]  80   5      0          5        1   130
#> [539,]  82   1      1          0        1    82
#> [540,]  79  10      0        180        1   150
#> [541,]  77   4      0        180        1    98
#> [542,]  81   1      0        108        0   129
#> [543,]  82  21      1          2        0   155
#> [544,]  79   4      0          4        1   121
#> [545,]  80   6      0          6        1   110
#> [546,]  83   4      0        103        0    97
#> [547,]  81  11      1          8        0   160
#> [548,]  80  11      1          8        0   170
#> [549,]  78  23      1         10        1   145
#> [550,]  82   8      1          1        0   128
#> [551,]  79   1      0        180        1   170
#> [552,]  80   7      1          0        1   146
#> [553,]  84   5      1          1        1    85
#> [554,]  83   8      0          8        0   115
#> [555,]  80   6      1          0        1   150
#> [556,]  79   0      1          0        1    96
#> [557,]  85   4      0        180        0    90
#> [558,]  81   2      1          1        0   198
#> [559,]  83   2      0          2        1   155
#> [560,]  82   6      0        128        1   100
#> [561,]  84   4      0        167        0   198
#> [562,]  80   3      1          1        1   230
#> [563,]  82  23      1          0        0   110
#> [564,]  84   5      0        180        1   203
#> [565,]  84   4      0          4        1    85
#> [566,]  81   1      0          1        1   150
#> [567,]  83   3      0        180        0   174
#> [568,]  81   4      0         90        1   138
#> [569,]  79   9      1          8        0   150
#> [570,]  85   3      1          2        1   160
#> [571,]  84   4      0         89        1   129
#> [572,]  80   6      0         71        1   189
#> [573,]  83   1      0          1        1   100
#> [574,]  82  19      0         19        0   120
#> [575,]  79  14      1          0        0   110
#> [576,]  83   3      0        114        0    98
#> [577,]  83   2      0        154        0   130
#> [578,]  82   0      0          2        1   100
#> [579,]  85   9      1          6        1   160
#> [580,]  81   4      0          4        0   175
#> [581,]  84  15      1         13        1   110
#> [582,]  81  12      0         12        1   163
#> [583,]  82  16      1          8        0   103
#> [584,]  81   4      0          4        0   160
#> [585,]  83  12      1          2        1   170
#> [586,]  81  19      1         14        0   120
#> [587,]  86   3      0          3        1   140
#> [588,]  84   3      0        180        1   120
#> [589,]  81  13      0        180        0   152
#> [590,]  81   2      1          0        1   118
#> [591,]  81   4      0        180        0   160
#> [592,]  83   9      0        180        1   149
#> [593,]  83   4      0          4        0   130
#> [594,]  82  14      1         11        1   103
#> [595,]  86   6      1          0        1   140
#> [596,]  84  16      0         70        1   150
#> [597,]  83  19      0         43        0   150
#> [598,]  84   3      1          2        0   125
#> [599,]  83  10      1          0        1   190
#> [600,]  86   2      0        180        1   169
#> [601,]  88  14      1          3        1   130
#> [602,]  84   3      0          3        1   121
#> [603,]  83  13      1         12        0   170
#> [604,]  84   7      1          2        0   148
#> [605,]  87   2      0        180        0   113
#> [606,]  84   9      0         92        1   110
#> [607,]  86   4      0         38        1   122
#> [608,]  86  13      0        177        0   163
#> [609,]  85   3      0          3        1   113
#> [610,]  86   6      0          6        1   117
#> [611,]  84  13      0         62        1   100
#> [612,]  86   6      1          1        0   112
#> [613,]  88   4      0          4        0   100
#> [614,]  88   4      0          4        1   115
#> [615,]  85  22      0         22        1   184
#> [616,]  86   6      0         46        0   173
#> [617,]  88   3      0        115        0   110
#> [618,]  88   2      0        180        1    68
#> [619,]  83   3      0          3        1   130
#> [620,]  87   8      0          8        1   157
#> [621,]  86  15      1          8        1   109
#> [622,]  88   4      0          4        0    86
#> [623,]  89   4      0          4        1   153
#> [624,]  87   6      0        180        1   110
#> [625,]  87   1      0          1        0   170
#> [626,]  84   8      0        180        1   119
#> [627,]  84   2      0        110        1   174
#> [628,]  87  15      1          9        1   138
#> [629,]  84   0      0        180        1   136
#> [630,]  89  10      0         46        1   170
#> [631,]  88   1      0          1        0   135
#> [632,]  86   4      0        180        1   145
#> [633,]  91   8      0          8        0   100
#> [634,]  91  10      0        145        0   135
#> [635,]  86   3      1          0        1    80
#> [636,]  88   7      0         24        0   119
#> [637,]  88   8      0         50        1   154
#> [638,]  90  11      1         10        1   186
#> [639,]  87   6      0        126        1   168
#> [640,]  86  10      0        180        1   137
#> [641,]  86   9      1          7        0   130
#> [642,]  90   4      1          0        0   121
#> [643,]  91   1      0          1        1    74
#> [644,]  88   3      1          2        0   159
#> [645,]  89   3      1          1        1   160
#> [646,]  92   1      0          1        1   167
#> [647,]  88   5      0        158        0   100
#> [648,]  87   7      0         74        1   105
#> [649,]  89   2      0        168        0   118
#> [650,]  91   5      0        169        1   176
#> [651,]  89  52      0         52        1   130
#> [652,]  92   7      0          7        1   110
#> [653,]  89   4      0          4        1   159
#> [654,]  91   0      0          0        0     0
#> [655,]  90  18      0        180        0   188
#> [656,]  90  19      1         11        1   129
#> [657,]  94   6      0         50        0    78
#> [658,]  93   8      0        179        1   110
#> [659,]  92   4      0         76        1   149
#> [660,]  91   1      0        180        0   158
#> [661,]  96   3      0         12        1    97
#> [662,]  95   8      1          5        1   150
#> [663,]  91  12      0         53        1   212
#> [664,]  91   7      0          7        0   135
#> [665,]  93   0      1          0        1   122
#> [666,]  92   5      0         69        0   139
#> [667,]  92   2      0          2        0   112
#> [668,]  93   4      0        180        1   135
#> [669,]  96   3      1          0        1   104
#> [670,]  96  15      1          0        1   140
#> 
#> $y
#>   [1]   5.0+ 180.0+   5.0+ 180.0+   2.0+ 180.0+ 180.0+ 115.0  180.0+  12.0 
#>  [11] 180.0+ 180.0+ 180.0+   2.0+ 180.0+   5.0+   2.0+ 180.0+ 180.0+ 180.0+
#>  [21]   3.0  180.0+ 180.0+ 180.0+   2.0+ 180.0+ 155.0+ 180.0+ 180.0+ 180.0+
#>  [31] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   6.0+ 180.0+ 180.0+
#>  [41] 180.0+   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 177.0+ 180.0+ 180.0+ 180.0+
#>  [51] 180.0+ 180.0+ 180.0+ 180.0+   7.0  180.0+ 180.0+ 180.0+ 179.0+ 180.0+
#>  [61] 179.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [71] 180.0+  36.0  180.0+ 180.0+ 180.0+ 180.0+  77.0  180.0+ 180.0+   5.0 
#>  [81] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  85.0  180.0+ 180.0+ 166.0+ 180.0+
#>  [91]  16.0+ 152.0+   7.0+   6.0+ 180.0+ 180.0+  13.0+ 171.0+ 180.0+ 174.0+
#> [101]  28.0    6.0+   1.0    9.0+ 180.0+ 180.0+ 180.0+   7.0+ 180.0+ 180.0+
#> [111] 180.0+ 180.0+  16.0+  16.0   15.0+ 180.0+  12.0+ 134.0+   8.0    2.0 
#> [121]   3.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [131]   0.5  180.0+ 180.0+ 180.0+ 171.0+  15.0  166.0+   5.0+ 180.0+   4.0+
#> [141] 147.0+ 180.0+ 180.0+ 180.0+   5.0+ 180.0+   4.0+ 180.0+   9.0+   1.0 
#> [151] 180.0+ 180.0+   2.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [161] 180.0+ 180.0+  64.0  180.0+ 180.0+ 180.0+ 180.0+ 171.0+   1.0  180.0+
#> [171] 180.0+ 180.0+  10.0+ 180.0+  17.0   45.0    3.0+  19.0  180.0+ 180.0+
#> [181] 172.0+ 172.0+  24.0  180.0+ 180.0+   1.0+ 180.0+ 180.0+  77.0   13.0+
#> [191]   8.0+ 180.0+ 180.0+  94.0  180.0+ 169.0    7.0    7.0+   6.0  180.0+
#> [201] 180.0+ 180.0+   3.0+ 180.0+  30.0  180.0+ 180.0+  13.0+   5.0   18.0 
#> [211]   5.0+ 180.0+ 180.0+   1.0  180.0+ 180.0+ 180.0+   4.0+ 180.0+  84.0 
#> [221]   7.0+ 180.0+ 180.0+   1.0  180.0+   3.0+   6.0+ 180.0+ 180.0+ 180.0+
#> [231]  14.0+   3.0+  88.0   12.0  180.0+ 180.0+ 180.0+ 180.0+   0.5  180.0+
#> [241] 180.0+  12.0+ 180.0+ 180.0+ 180.0+ 180.0+  12.0  180.0+   9.0    3.0 
#> [251] 180.0+ 180.0+  18.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   3.0+   2.0+
#> [261]  15.0  180.0+ 180.0+  13.0  179.0+ 166.0+  14.0+   3.0    0.5+ 180.0+
#> [271] 180.0+ 175.0+ 180.0+   7.0+   8.0    5.0   16.0  180.0+   1.0  180.0+
#> [281] 180.0+ 123.0+   1.0+  11.0+  79.0   80.0   15.0+   4.0+ 180.0+ 180.0+
#> [291] 180.0+ 180.0+ 180.0+ 180.0+   8.0+   3.0   10.0  180.0+ 180.0+  19.0+
#> [301]  12.0   99.0  180.0+ 179.0+ 180.0+   0.5   18.0  180.0+ 180.0+ 152.0+
#> [311] 180.0+ 180.0+   1.0   18.0+ 101.0    4.0    7.0+ 150.0  180.0+ 180.0+
#> [321] 180.0+ 180.0+ 171.0  180.0+   6.0  180.0+   0.5  180.0+ 180.0+ 180.0+
#> [331]  14.0+ 180.0+   7.0+   2.0   45.0    5.0+ 180.0+ 180.0+ 180.0+   7.0 
#> [341] 180.0+ 180.0+  18.0  172.0+ 180.0+ 180.0+   7.0  180.0+   8.0+  13.0+
#> [351] 123.0  180.0+ 180.0+  19.0  180.0+   1.0    1.0   60.0   76.0  180.0+
#> [361] 180.0+   7.0+   7.0+ 124.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [371] 180.0+ 180.0+  28.0  180.0+ 180.0+  16.0+ 180.0+ 180.0+  16.0+ 180.0+
#> [381]   6.0  180.0+ 180.0+  15.0  180.0+ 180.0+ 180.0+   3.0+ 180.0+  20.0 
#> [391] 180.0+   0.5  180.0+   8.0    3.0   87.0    4.0+ 180.0+ 180.0+   3.0 
#> [401] 180.0+ 175.0    2.0   14.0+ 180.0+ 180.0+   1.0  180.0+ 159.0  180.0+
#> [411]  10.0    1.0   13.0    4.0+ 180.0+  10.0    1.0   57.0  180.0+  11.0 
#> [421] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 177.0+ 180.0+   4.0+ 180.0+ 180.0+
#> [431] 180.0+   3.0  180.0+ 180.0+ 180.0+ 180.0+   6.0  180.0+ 180.0+  85.0 
#> [441] 180.0+  17.0+ 180.0+   7.0    0.5  180.0+  12.0  180.0+  46.0    4.0 
#> [451]   1.0  180.0+  52.0  180.0+  12.0  180.0+ 180.0+ 180.0+   8.0  180.0+
#> [461] 180.0+   1.0   12.0  180.0+   7.0+  79.0  168.0+ 180.0+ 180.0+ 176.0+
#> [471]  47.0  180.0+  11.0    7.0    8.0+   7.0   32.0  180.0+ 180.0+ 172.0 
#> [481] 119.0   12.0  180.0+   8.0    1.0  180.0+ 180.0+   4.0+   2.0   11.0 
#> [491] 152.0+   3.0   24.0   32.0    6.0  180.0+ 180.0+   1.0   11.0    4.0 
#> [501]   4.0   10.0+   6.0  180.0+   1.0  171.0    1.0   43.0    3.0    6.0 
#> [511] 138.0   71.0    8.0   40.0   59.0   17.0  161.0   10.0+  93.0  118.0 
#> [521] 173.0  180.0+   7.0+  22.0   15.0+   3.0  171.0+ 166.0+   1.0    3.0+
#> [531]  10.0  180.0+  85.0   10.0    6.0+   6.0  180.0+   5.0    1.0  180.0+
#> [541] 180.0+ 108.0  180.0+   4.0    6.0  103.0  180.0+ 169.0   70.0  180.0+
#> [551] 180.0+   7.0+ 180.0+   8.0+ 180.0+   0.5  180.0+ 180.0+   2.0  128.0 
#> [561] 167.0    3.0+  62.0  180.0+   4.0    1.0  180.0+  90.0  180.0+ 180.0+
#> [571]  89.0   71.0    1.0   19.0  180.0+ 114.0  154.0    2.0  180.0+   4.0+
#> [581] 180.0+  12.0   16.0+   4.0+  77.0  180.0+   3.0  180.0+ 180.0+ 180.0+
#> [591] 180.0+ 180.0+   4.0+ 174.0    6.0   70.0   43.0  180.0+ 180.0+ 180.0+
#> [601]  14.0    3.0   13.0  180.0+ 180.0+  92.0   38.0  177.0    3.0+   6.0+
#> [611]  62.0    6.0+   4.0+   4.0   22.0   46.0  115.0  180.0+   3.0+   8.0+
#> [621] 180.0+   4.0    4.0  180.0+   1.0+ 180.0+ 110.0  180.0+ 180.0+  46.0 
#> [631]   1.0+ 180.0+   8.0  145.0    3.0   24.0   50.0   11.0  126.0  180.0+
#> [641] 180.0+   4.0    1.0   75.0    3.0+   1.0  158.0   74.0  168.0  169.0 
#> [651]  52.0    7.0    4.0    0.5  180.0+ 180.0+  50.0  179.0+  76.0  180.0+
#> [661]  12.0    8.0   53.0    7.0+   0.5   69.0    2.0  180.0+   3.0   15.0+
#> 
#> $weights
#> NULL
#> 
#> $offset
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8214551