GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.coxnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.coxnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Caution: This learner is different to learners calling glmnet::cv.glmnet() in that it does not use the internal optimization of parameter lambda. Instead, lambda needs to be tuned by the user (e.g., via mlr3tuning). When lambda is tuned, the glmnet will be trained for each tuning iteration. While fitting the whole path of lambdas would be more efficient, as is done by default in glmnet::glmnet(), tuning/selecting the parameter at prediction time (using parameter s) is currently not supported in mlr3 (at least not in efficient manner). Tuning the s parameter is, therefore, currently discouraged.

When the data are i.i.d. and efficiency is key, we recommend using the respective auto-tuning counterpart in mlr_learners_surv.cv_glmnet(). However, in some situations this is not applicable, usually when data are imbalanced or not i.i.d. (longitudinal, time-series) and tuning requires custom resampling strategies (blocked design, stratification).

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exact	logical	FALSE	TRUE, FALSE	-

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

newoffset	untyped	-		-

nlambda	integer	100		\([1, \infty)\)

offset	untyped	NULL		-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	0.01		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.logistic	character	Newton	Newton, modified.Newton	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvGlmnet

Methods

Public methods

LearnerSurvGlmnet$new()

LearnerSurvGlmnet$selected_features()

LearnerSurvGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.glmnet")
print(learner)
#> <LearnerSurvGlmnet:surv.glmnet>: Regularized Generalized Linear Model
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3proba, mlr3extralearners, glmnet
#> * Predict Types:  [crank], distr, lp
#> * Feature Types: logical, integer, numeric
#> * Properties: selected_features, weights

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#>    Df  %Dev   Lambda
#> 1   0  0.00 0.190300
#> 2   1  0.66 0.173400
#> 3   2  1.54 0.158000
#> 4   2  2.47 0.144000
#> 5   2  3.28 0.131200
#> 6   3  4.12 0.119500
#> 7   3  6.99 0.108900
#> 8   3  9.06 0.099230
#> 9   3 10.63 0.090410
#> 10  3 11.86 0.082380
#> 11  3 12.85 0.075060
#> 12  4 13.67 0.068390
#> 13  4 14.38 0.062320
#> 14  5 14.97 0.056780
#> 15  5 15.47 0.051740
#> 16  5 15.89 0.047140
#> 17  5 16.25 0.042950
#> 18  5 16.54 0.039140
#> 19  5 16.79 0.035660
#> 20  5 17.00 0.032490
#> 21  5 17.18 0.029610
#> 22  5 17.33 0.026980
#> 23  6 17.46 0.024580
#> 24  6 17.57 0.022400
#> 25  6 17.67 0.020410
#> 26  6 17.75 0.018590
#> 27  6 17.82 0.016940
#> 28  6 17.87 0.015440
#> 29  6 17.92 0.014070
#> 30  6 17.96 0.012820
#> 31  6 17.99 0.011680
#> 32  6 18.02 0.010640
#> 33  6 18.05 0.009694
#> 34  6 18.07 0.008833
#> 35  6 18.08 0.008049
#> 36  6 18.10 0.007334
#> 37  6 18.11 0.006682
#> 38  6 18.12 0.006088
#> 39  6 18.13 0.005548
#> 40  6 18.13 0.005055
#> 41  6 18.14 0.004606
#> 42  6 18.14 0.004197
#> 43  6 18.15 0.003824
#> 44  6 18.15 0.003484
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  28   9      0        180        1   107
#>   [2,]  33   2      0          2        0   150
#>   [3,]  35   5      1          2        0   172
#>   [4,]  35  10      1          9        0   106
#>   [5,]  34   5      0          5        0   120
#>   [6,]  35   2      0        180        0   121
#>   [7,]  37   9      0        180        1   151
#>   [8,]  38  13      1          0        1   161
#>   [9,]  36   1      0        180        1   155
#>  [10,]  35   0      0        180        1   119
#>  [11,]  36   5      1          0        1   115
#>  [12,]  33   6      1          1        1   115
#>  [13,]  38  16      1         10        0   160
#>  [14,]  40  12      1          9        0   153
#>  [15,]  42   3      1          1        1   130
#>  [16,]  37   1      1          0        1   146
#>  [17,]  42   2      0        180        1   100
#>  [18,]  38   5      1          3        0   125
#>  [19,]  40   6      0        180        1   138
#>  [20,]  40  11      1         10        1   120
#>  [21,]  42   2      0        180        0   100
#>  [22,]  43   3      1          0        1   100
#>  [23,]  41   2      1          1        0   166
#>  [24,]  40   1      1          0        1   145
#>  [25,]  42   4      0        180        0   162
#>  [26,]  40   3      1          1        0   170
#>  [27,]  42  12      1         10        1   170
#>  [28,]  43   2      1          1        1   116
#>  [29,]  42   2      0        180        1   124
#>  [30,]  41  10      1          8        0   150
#>  [31,]  41  13      1          1        0   140
#>  [32,]  45   6      0        180        1   170
#>  [33,]  41   5      1          4        1   141
#>  [34,]  43   2      0        180        1   140
#>  [35,]  46   2      1          1        0   126
#>  [36,]  48  15      0        180        1   160
#>  [37,]  46   7      1          2        0   166
#>  [38,]  43  29      0        180        1   180
#>  [39,]  45   4      1          0        0   124
#>  [40,]  43  10      0        180        0   185
#>  [41,]  47   6      1          0        1   116
#>  [42,]  47   4      1          3        1   160
#>  [43,]  43   3      1          0        1   124
#>  [44,]  45   8      1          0        1   117
#>  [45,]  49   5      0         73        1   136
#>  [46,]  46   6      1          0        1   100
#>  [47,]  44   4      1          0        1   114
#>  [48,]  44   9      1          8        1   135
#>  [49,]  45   5      0        180        1   190
#>  [50,]  46   5      1          3        0   130
#>  [51,]  46   4      0        180        1   121
#>  [52,]  46  15      0        180        1   120
#>  [53,]  45   9      1          0        1   145
#>  [54,]  47   3      1          1        1   120
#>  [55,]  48   3      0        180        0   154
#>  [56,]  47   5      1          3        1   130
#>  [57,]  47   9      1          6        0   170
#>  [58,]  46   3      1          0        1   119
#>  [59,]  49   4      0        180        0   117
#>  [60,]  47  10      0         10        1   140
#>  [61,]  48   2      1          0        0   184
#>  [62,]  50   4      1          1        0   125
#>  [63,]  50   6      1          2        1   140
#>  [64,]  49   7      1          7        1   110
#>  [65,]  46   9      1          9        1   122
#>  [66,]  50   7      0        180        1   110
#>  [67,]  51   1      0          1        1   145
#>  [68,]  47   2      0        180        0   150
#>  [69,]  49  23      0        179        1   112
#>  [70,]  46   6      1          0        1   156
#>  [71,]  52   2      0        180        1   170
#>  [72,]  50   7      1          0        1    92
#>  [73,]  51   3      1          2        0   113
#>  [74,]  50   1      1          0        0   150
#>  [75,]  49   7      1          4        1    90
#>  [76,]  47   6      0        180        1   162
#>  [77,]  52   2      0        180        0   155
#>  [78,]  46   1      1          1        0   145
#>  [79,]  50   4      1          1        0   150
#>  [80,]  48   7      1          0        1   110
#>  [81,]  53   8      0         36        1   160
#>  [82,]  47   2      1          1        0   110
#>  [83,]  49  15      0        180        1   160
#>  [84,]  53   5      0        180        1   140
#>  [85,]  54  17      1         12        1   102
#>  [86,]  53   7      1          0        0   199
#>  [87,]  54   6      1          3        0   129
#>  [88,]  50  10      1          6        0   122
#>  [89,]  50  14      1         13        0   170
#>  [90,]  53   8      1          7        0   160
#>  [91,]  48   3      1          2        0   150
#>  [92,]  51  25      1          1        0   202
#>  [93,]  49   5      1          2        1   150
#>  [94,]  53   4      0          4        0   140
#>  [95,]  48   6      0        180        0   160
#>  [96,]  48  11      1         10        0   120
#>  [97,]  53   4      1          0        1   156
#>  [98,]  54   9      1          0        1   138
#>  [99,]  49  16      0         16        0   125
#> [100,]  55   3      1          1        0   150
#> [101,]  54  23      1         10        0   131
#> [102,]  52   7      1          2        0   154
#> [103,]  55   6      1          2        1   114
#> [104,]  54   9      1          1        0   130
#> [105,]  55   4      1          2        0   150
#> [106,]  51  13      1         11        0   145
#> [107,]  50   5      1          4        1   150
#> [108,]  52   4      0        180        0   183
#> [109,]  49   6      1          0        1   130
#> [110,]  49   1      0          1        1   110
#> [111,]  50   7      1          0        1   127
#> [112,]  56   4      1          1        1   130
#> [113,]  52   5      0        175        1   117
#> [114,]  55   1      0        180        0   127
#> [115,]  54   1      0        180        0   162
#> [116,]  56   3      0        180        1   193
#> [117,]  56   2      0        180        0   132
#> [118,]  55   5      1          4        1   120
#> [119,]  52   8      0        180        0   119
#> [120,]  53  18      1          9        1   150
#> [121,]  54   3      0        180        1   180
#> [122,]  55   6      0        180        0   170
#> [123,]  52  16      0         16        0   152
#> [124,]  53  10      1          9        0   172
#> [125,]  52  16      1         14        0   170
#> [126,]  53  15      0         15        1    90
#> [127,]  55   6      0        180        1   100
#> [128,]  55   2      0        134        1   140
#> [129,]  54   3      0        180        0   128
#> [130,]  56   3      0          8        1   139
#> [131,]  57   3      0          3        0   120
#> [132,]  54   2      1          1        0   135
#> [133,]  52   9      1          3        0   170
#> [134,]  57   5      1          3        1   138
#> [135,]  57   1      0        180        1   156
#> [136,]  57   1      0          1        1   100
#> [137,]  56   4      1          0        1   140
#> [138,]  52   2      0        180        0   140
#> [139,]  52  15      1         14        0   130
#> [140,]  56  14      1         11        0   130
#> [141,]  57  10      0        180        1   170
#> [142,]  58   8      0          8        1   130
#> [143,]  54   5      0        180        1   108
#> [144,]  55   3      1          1        1   156
#> [145,]  57   0      0          0        1   150
#> [146,]  53  15      1         10        1   130
#> [147,]  54  17      1          8        1   227
#> [148,]  55  13      0        166        1   140
#> [149,]  56   5      0          5        1   150
#> [150,]  54  23      1          8        0   120
#> [151,]  57   4      1          2        1   185
#> [152,]  53   4      0        147        1   145
#> [153,]  57  11      1         10        1   129
#> [154,]  55   3      1          2        0   140
#> [155,]  54   7      1          0        1   141
#> [156,]  56   4      0          4        0   164
#> [157,]  59  15      1         10        0   140
#> [158,]  58   9      1          0        1   180
#> [159,]  58   1      1          1        1   200
#> [160,]  56   7      1          5        1   120
#> [161,]  55   2      0          2        0   106
#> [162,]  59   9      1          1        1   125
#> [163,]  57   1      0        180        0   148
#> [164,]  60  11      1          9        0   106
#> [165,]  58   4      1          0        1   160
#> [166,]  57   2      0          2        1   120
#> [167,]  60   5      1          1        0   138
#> [168,]  57   5      0        180        1   130
#> [169,]  58  11      1          9        1   124
#> [170,]  55   5      1          0        1   160
#> [171,]  57  10      1          9        0   103
#> [172,]  59   5      0        180        1   155
#> [173,]  58  26      1          0        1   189
#> [174,]  61   9      0          9        1   160
#> [175,]  58   4      1          3        0   120
#> [176,]  60   0      1          0        1    80
#> [177,]  59   2      1          1        0   140
#> [178,]  58   8      0        161        1   140
#> [179,]  58  14      1          6        0   190
#> [180,]  61   4      1          3        0   151
#> [181,]  61   3      1          2        1   102
#> [182,]  58   1      0          1        1   100
#> [183,]  61  20      1         13        0   130
#> [184,]  57  13      1         10        0   110
#> [185,]  58  10      0         10        1   150
#> [186,]  57   4      1          3        0   138
#> [187,]  61   3      0         17        0   143
#> [188,]  56  13      1          6        1   158
#> [189,]  56  18      1         11        1   165
#> [190,]  55   4      1          3        1   160
#> [191,]  58  11      0        172        1   135
#> [192,]  55   9      1          7        1   135
#> [193,]  56   8      1          8        0   120
#> [194,]  61  13      1         12        1   130
#> [195,]  59  11      1          8        1   190
#> [196,]  57   1      0          1        0   126
#> [197,]  58   5      1          1        1   135
#> [198,]  61   8      0         77        0   120
#> [199,]  58   8      1          5        0   152
#> [200,]  61  28      1          7        0   133
#> [201,]  58   8      1          3        1   150
#> [202,]  57   7      0        169        0   180
#> [203,]  59  13      1          2        0   198
#> [204,]  62   4      1          0        0   160
#> [205,]  60  17      1          8        1   140
#> [206,]  62   4      1          3        0   173
#> [207,]  59   1      0        180        0   155
#> [208,]  61   5      0          5        1   110
#> [209,]  57  18      1          9        1    93
#> [210,]  58   7      0        180        1   150
#> [211,]  63   1      0        180        1   130
#> [212,]  61   7      0        180        0   135
#> [213,]  63   4      0        180        1   190
#> [214,]  64   4      0        180        0   130
#> [215,]  63   4      1          1        0   155
#> [216,]  60  18      1         13        0   132
#> [217,]  59   8      0        180        1   140
#> [218,]  58   9      1          9        0   110
#> [219,]  59   1      0         22        1   162
#> [220,]  58   2      0        180        0   127
#> [221,]  60   7      1          5        1   141
#> [222,]  60   7      0          7        0   140
#> [223,]  59   5      1          1        0   148
#> [224,]  65  13      0        180        1   100
#> [225,]  63   1      0          1        0   162
#> [226,]  63   1      0          1        0   130
#> [227,]  62   6      0        180        0   170
#> [228,]  61  15      1         13        0   170
#> [229,]  59   4      0          4        0   149
#> [230,]  60   3      0          3        0   168
#> [231,]  64  10      1          9        0   160
#> [232,]  62   6      0          6        0   120
#> [233,]  59  10      0        180        1   130
#> [234,]  61   6      1          1        1   117
#> [235,]  66   1      1          0        1   120
#> [236,]  65  36      1         11        0   140
#> [237,]  63   4      1          3        0   162
#> [238,]  66   3      1          1        0   127
#> [239,]  64  32      1          9        1   160
#> [240,]  66   5      1          0        1   110
#> [241,]  65   8      1          0        0   168
#> [242,]  64   0      0          0        1    90
#> [243,]  64  21      1         10        0   190
#> [244,]  61  12      1         11        0   154
#> [245,]  64   9      0        180        0   150
#> [246,]  63  16      1          7        1   110
#> [247,]  64   7      0        180        1   120
#> [248,]  66   6      1          1        1   130
#> [249,]  62   3      1          1        1   199
#> [250,]  65   6      0          9        0   112
#> [251,]  63   5      1          4        0   170
#> [252,]  63   2      1          1        0   180
#> [253,]  62  13      1         11        0   180
#> [254,]  66  18      1          5        0   142
#> [255,]  61  14      1          5        0   140
#> [256,]  63   3      1          2        0   120
#> [257,]  63   2      1          0        0   140
#> [258,]  64  19      1          8        1   160
#> [259,]  65   8      1          0        1   140
#> [260,]  67   6      0        180        1   170
#> [261,]  65  15      1         11        1   160
#> [262,]  64  13      1         12        1   150
#> [263,]  66   7      1          0        1   115
#> [264,]  64  14      1         13        1   150
#> [265,]  64   0      0          0        1   148
#> [266,]  67   4      1          3        0   130
#> [267,]  66   3      1          0        1   135
#> [268,]  66   6      1          0        1   140
#> [269,]  65   2      1          1        1   170
#> [270,]  68   1      0        180        1   166
#> [271,]  63   7      1          0        0   162
#> [272,]  68   5      0          5        1    90
#> [273,]  64   1      0          1        1   120
#> [274,]  63   8      1          1        1   162
#> [275,]  65  18      1          3        0   120
#> [276,]  63  10      0         18        1   130
#> [277,]  67  11      0         11        0   150
#> [278,]  68  14      0         79        0   172
#> [279,]  66  12      1         10        1   150
#> [280,]  65  15      1         12        1   150
#> [281,]  66  11      1          0        0   100
#> [282,]  69  12      0         15        1   140
#> [283,]  66  15      1         13        1   160
#> [284,]  69  21      1         10        0   180
#> [285,]  69   6      0        180        1   100
#> [286,]  63   8      0        180        1   120
#> [287,]  68  14      1         13        1   140
#> [288,]  67   1      0        180        1   160
#> [289,]  68  10      1         10        1   150
#> [290,]  67   7      1          4        1   130
#> [291,]  63   2      1          0        0    99
#> [292,]  67   2      0        180        0   184
#> [293,]  65   6      0          6        0    80
#> [294,]  65  10      1          1        1   148
#> [295,]  66  19      1         12        1   150
#> [296,]  67  12      1         12        0   160
#> [297,]  69   6      0         99        1   140
#> [298,]  65   4      1          1        0   130
#> [299,]  70  15      1         12        1   132
#> [300,]  64   4      0        180        1   140
#> [301,]  67   2      0         18        0   131
#> [302,]  66   7      1          5        1   131
#> [303,]  66   4      0        180        0   177
#> [304,]  69   4      1          3        1   150
#> [305,]  69  17      1         10        0   140
#> [306,]  69   8      0         93        0   140
#> [307,]  64  21      0         21        1   155
#> [308,]  65   1      0          1        1   120
#> [309,]  68   4      0          4        1   190
#> [310,]  71   3      0          5        0   112
#> [311,]  70   7      1          0        1   190
#> [312,]  68   7      0        150        0   210
#> [313,]  71  20      1          0        1   160
#> [314,]  67   2      0        180        0   128
#> [315,]  66   9      1          3        1   151
#> [316,]  66   1      1          1        1   165
#> [317,]  70   4      1          0        1   180
#> [318,]  66   4      0        180        0   130
#> [319,]  67   6      1          4        0   130
#> [320,]  68  18      1         14        1   170
#> [321,]  69   0      0          0        1   148
#> [322,]  65   2      0        180        0   130
#> [323,]  69   8      0        180        1   180
#> [324,]  71   7      0          7        0   230
#> [325,]  66   2      0          2        1   228
#> [326,]  69   3      0          3        1   130
#> [327,]  70  22      1         13        0   103
#> [328,]  69   8      1          5        1   195
#> [329,]  69   6      1          4        1   174
#> [330,]  72   7      0          7        1   110
#> [331,]  69   8      1          7        1   108
#> [332,]  68  18      0         18        1   100
#> [333,]  67  14      0        172        1   140
#> [334,]  66   2      0        180        0   130
#> [335,]  67   7      1          4        0   122
#> [336,]  69   8      1          2        0   121
#> [337,]  70   3      0        123        0   130
#> [338,]  70   9      0        180        1   142
#> [339,]  72   5      1          4        0   170
#> [340,]  67  22      1          1        1   140
#> [341,]  67  12      1          8        0   120
#> [342,]  69   1      0          1        1   110
#> [343,]  67   1      0          1        1    60
#> [344,]  69   5      0         76        0   120
#> [345,]  67   8      1          0        1   130
#> [346,]  72  13      1         11        1   195
#> [347,]  68  10      1          8        1   160
#> [348,]  66  24      1         13        0   130
#> [349,]  72  30      1          0        1   145
#> [350,]  73  20      1          0        1   170
#> [351,]  70  11      0        180        1   210
#> [352,]  72  19      1          8        0   120
#> [353,]  72  12      1         10        0   170
#> [354,]  67   8      0        180        1   170
#> [355,]  67   9      0        180        0   158
#> [356,]  70   5      0        180        0   150
#> [357,]  67   4      1          1        0   134
#> [358,]  72   6      1          5        0   115
#> [359,]  68  23      0        180        1   220
#> [360,]  69   3      0        180        0   220
#> [361,]  72   5      0         28        0   120
#> [362,]  71   5      0        180        0   191
#> [363,]  73   6      0        180        1   117
#> [364,]  69  16      1         10        1   140
#> [365,]  68   7      0        180        1   130
#> [366,]  72  16      1          1        1   130
#> [367,]  70   4      0        180        0   180
#> [368,]  69   1      1          0        0   155
#> [369,]  73   6      1          0        1   270
#> [370,]  72   8      1          1        1   150
#> [371,]  71   2      1          0        1   180
#> [372,]  73   7      0          7        1   140
#> [373,]  70  13      1          9        0   100
#> [374,]  72   6      0        180        1   130
#> [375,]  73   0      0        180        1   161
#> [376,]  74   8      1          0        1    85
#> [377,]  73   4      0        180        1   154
#> [378,]  71  15      1         11        0   165
#> [379,]  74  20      0         20        1   180
#> [380,]  68   9      0        180        1   120
#> [381,]  71  20      1         10        0   140
#> [382,]  74   0      1          0        1    90
#> [383,]  73   3      1          0        1   136
#> [384,]  71  17      1         11        0   160
#> [385,]  71   3      1          2        1   190
#> [386,]  73  10      1          8        0   106
#> [387,]  69  12      1          1        1   149
#> [388,]  74   4      0          4        0   120
#> [389,]  73   4      0         58        1   160
#> [390,]  72   5      1          3        1   160
#> [391,]  72  15      1          0        1   150
#> [392,]  71   7      1          2        0   143
#> [393,]  72   8      1          0        1   140
#> [394,]  74   3      0          3        1   150
#> [395,]  71  13      1          8        0   121
#> [396,]  69   2      1          1        1    80
#> [397,]  70   4      1          0        1   140
#> [398,]  71  14      1         13        1   170
#> [399,]  74   7      1          0        1   117
#> [400,]  72  10      1          8        1   153
#> [401,]  69   7      0        180        1   144
#> [402,]  72  15      1         13        0   156
#> [403,]  70   8      0          8        0   120
#> [404,]  71  10      1          9        1   120
#> [405,]  72  10      1          9        1   160
#> [406,]  73  10      1         10        1   120
#> [407,]  75   9      1          7        0   140
#> [408,]  75  13      1          1        1   130
#> [409,]  71   4      0          4        0   134
#> [410,]  72  15      1         12        1   120
#> [411,]  70   7      1          4        0   184
#> [412,]  73  10      0        180        0   162
#> [413,]  70   3      0          3        0   150
#> [414,]  73   5      1          3        1   112
#> [415,]  76  25      1         12        1   170
#> [416,]  73  12      1         12        1   140
#> [417,]  72   2      0        180        0   120
#> [418,]  75   1      0        180        1   140
#> [419,]  72   4      1          0        1   197
#> [420,]  71   3      1          0        0   144
#> [421,]  73   5      0        180        0   126
#> [422,]  74  34      1          8        1   233
#> [423,]  76   3      1          0        1   120
#> [424,]  72   5      0        180        0   154
#> [425,]  76   5      0          5        1   130
#> [426,]  77   4      0          4        0   185
#> [427,]  75   3      1          1        0   180
#> [428,]  72   7      1          2        0   142
#> [429,]  71  16      0        180        0   140
#> [430,]  74   3      0          3        1   128
#> [431,]  76   1      0        180        0   114
#> [432,]  74   2      1          1        0   140
#> [433,]  74  19      1          4        1   200
#> [434,]  73   6      0          6        1   114
#> [435,]  75  23      1         14        1   110
#> [436,]  74   2      0        180        0   190
#> [437,]  72   4      0         85        1   120
#> [438,]  72   4      1          3        0   160
#> [439,]  76  17      1          0        1   200
#> [440,]  73   4      1          3        1   125
#> [441,]  75   4      1          0        1   122
#> [442,]  75   7      0          7        0   190
#> [443,]  75   0      0          0        1   130
#> [444,]  73  13      1         11        0   195
#> [445,]  75  12      0         12        1   160
#> [446,]  74   8      1          0        1   105
#> [447,]  76  13      1          8        1   148
#> [448,]  76   4      0          4        1   155
#> [449,]  75   1      0          1        1   125
#> [450,]  74   2      0        180        0   111
#> [451,]  73   1      0         52        1   105
#> [452,]  73   0      0        180        0   156
#> [453,]  72   5      0        180        0   120
#> [454,]  78  12      1         11        1   160
#> [455,]  76   5      0        180        0   185
#> [456,]  74  10      1          0        1   135
#> [457,]  76   5      1          0        1   167
#> [458,]  75   9      0        180        1   140
#> [459,]  73  33      1         12        1   175
#> [460,]  77   5      1          0        0   123
#> [461,]  73  10      1          9        0   146
#> [462,]  76  12      1         11        1   120
#> [463,]  73   7      1          0        0   174
#> [464,]  75   3      1          1        1   171
#> [465,]  75   6      0        180        0   150
#> [466,]  79  10      1          8        0   190
#> [467,]  74   2      1          0        1   130
#> [468,]  77   3      0        180        0   110
#> [469,]  76  29      0         47        0    90
#> [470,]  73   8      1          1        1   162
#> [471,]  78   7      0          7        1   133
#> [472,]  74  15      0        180        1   172
#> [473,]  79   6      0        180        0   170
#> [474,]  80  10      1          6        1   147
#> [475,]  78  13      1          5        0   130
#> [476,]  75  12      1          1        1   120
#> [477,]  78  15      0        180        1   270
#> [478,]  80   8      0          8        1   120
#> [479,]  75  13      1          6        0   150
#> [480,]  74  10      1          8        0   135
#> [481,]  79   4      0         80        0   145
#> [482,]  75   4      1          0        0   212
#> [483,]  77   2      1          0        1   143
#> [484,]  75  11      1          4        0   162
#> [485,]  76   7      0         29        1   150
#> [486,]  77  24      0         24        1   160
#> [487,]  79   8      0         32        1   120
#> [488,]  80   9      0         23        1   128
#> [489,]  80   6      0          6        1   150
#> [490,]  78   6      1          0        1   240
#> [491,]  76   3      1          0        1   140
#> [492,]  78  11      1          1        1   140
#> [493,]  79  11      0        180        0   160
#> [494,]  79   2      1          0        1   121
#> [495,]  78  14      1          0        1   140
#> [496,]  81   1      0          1        0   130
#> [497,]  78  11      1          8        1   118
#> [498,]  76   4      0          4        1   160
#> [499,]  79   4      0          4        1   125
#> [500,]  76  10      1          8        0   180
#> [501,]  76  12      1         10        1   127
#> [502,]  80   3      1          0        1   120
#> [503,]  78  11      0        180        1   135
#> [504,]  77  31      1          3        1   161
#> [505,]  76   1      0          1        1    90
#> [506,]  78   7      1          0        1   110
#> [507,]  77   7      0        180        1   170
#> [508,]  77   6      0          6        1   144
#> [509,]  79   4      1          0        1   120
#> [510,]  81   1      0        180        0   120
#> [511,]  80  15      1         12        1   150
#> [512,]  82   5      0          8        1   120
#> [513,]  80  40      1          0        1   138
#> [514,]  80  17      1         12        0   100
#> [515,]  76   7      0        161        0   151
#> [516,]  80  15      1          0        1    90
#> [517,]  79  28      0        164        0   100
#> [518,]  80   9      0        118        1   186
#> [519,]  80   6      0        173        1   160
#> [520,]  81   3      0        180        0   184
#> [521,]  81   2      0        175        0   172
#> [522,]  77  13      1          0        1   190
#> [523,]  79   3      0          3        1   101
#> [524,]  78  26      1          5        0   194
#> [525,]  76   1      0        166        0   131
#> [526,]  78  20      1          0        1   109
#> [527,]  80   1      0          1        0   100
#> [528,]  78   3      1          1        1   152
#> [529,]  77  10      1          8        1   130
#> [530,]  82   3      1          1        1   144
#> [531,]  77   5      0         85        0   188
#> [532,]  80   2      1          1        0   168
#> [533,]  79   6      0          6        0   152
#> [534,]  80   6      1          0        1   119
#> [535,]  78   2      0        180        0   148
#> [536,]  80   5      0          5        1   130
#> [537,]  82   1      1          0        1    82
#> [538,]  79  10      0        180        1   150
#> [539,]  77   4      0        180        1    98
#> [540,]  81   1      0        108        0   129
#> [541,]  78  12      0        180        0   134
#> [542,]  79   1      0        125        0   193
#> [543,]  82  21      1          2        0   155
#> [544,]  79   4      0          4        1   121
#> [545,]  80   6      0          6        1   110
#> [546,]  83   5      0        180        0   148
#> [547,]  79   7      1          6        0   130
#> [548,]  83   4      0        103        0    97
#> [549,]  81  11      1          8        0   160
#> [550,]  80  11      1          8        0   170
#> [551,]  78  23      1         10        1   145
#> [552,]  78   9      1          4        1   120
#> [553,]  82   8      1          1        0   128
#> [554,]  79   1      0        180        1   170
#> [555,]  81  15      0        180        1   140
#> [556,]  80   7      1          0        1   146
#> [557,]  84   5      1          1        1    85
#> [558,]  81  20      1          9        0   170
#> [559,]  83   8      0          8        0   115
#> [560,]  81  16      0         16        1   110
#> [561,]  80   6      1          0        1   150
#> [562,]  81   8      0        180        0   146
#> [563,]  80   8      1          7        0   160
#> [564,]  79   0      1          0        1    96
#> [565,]  81   2      1          1        0   198
#> [566,]  83   2      0          2        1   155
#> [567,]  82   6      0        128        1   100
#> [568,]  84   4      0        167        0   198
#> [569,]  82  23      1          0        0   110
#> [570,]  84   5      0        180        1   203
#> [571,]  84   4      0          4        1    85
#> [572,]  84   1      0         38        1   205
#> [573,]  81   4      0         90        1   138
#> [574,]  79   9      1          8        0   150
#> [575,]  85   3      1          2        1   160
#> [576,]  80   2      1          0        1   130
#> [577,]  79   4      0          4        1    60
#> [578,]  80   6      0         71        1   189
#> [579,]  82  19      0         19        0   120
#> [580,]  80  30      1         13        0   220
#> [581,]  83   9      0        180        0   198
#> [582,]  79  14      1          0        0   110
#> [583,]  81  14      1         12        1   128
#> [584,]  83   2      0        154        0   130
#> [585,]  82   0      0          2        1   100
#> [586,]  85   9      1          6        1   160
#> [587,]  81   4      0          4        0   175
#> [588,]  84  15      1         13        1   110
#> [589,]  81  12      0         12        1   163
#> [590,]  82  16      1          8        0   103
#> [591,]  82   5      1          0        1   146
#> [592,]  86  12      0        180        1   120
#> [593,]  82  15      1          0        0   183
#> [594,]  80   2      0         88        0   135
#> [595,]  81  16      1          9        0   180
#> [596,]  86   3      0          3        1   140
#> [597,]  85   3      0          3        1   118
#> [598,]  81   2      1          0        1   118
#> [599,]  81   4      0        180        0   160
#> [600,]  83   9      0        180        1   149
#> [601,]  82   1      0        180        1   193
#> [602,]  83   4      0          4        0   130
#> [603,]  87   2      0          5        1   137
#> [604,]  86  12      1          0        1   132
#> [605,]  86   6      1          0        1   140
#> [606,]  84  16      0         70        1   150
#> [607,]  84   3      1          2        0   125
#> [608,]  83  10      1          0        1   190
#> [609,]  86   2      0        180        1   169
#> [610,]  88  14      1          3        1   130
#> [611,]  84   3      0          3        1   121
#> [612,]  83  13      1         12        0   170
#> [613,]  84   9      0         92        1   110
#> [614,]  84   3      0        180        1   170
#> [615,]  86   4      0         38        1   122
#> [616,]  86  13      0        177        0   163
#> [617,]  86   6      0          6        1   117
#> [618,]  84  13      0         62        1   100
#> [619,]  86   6      1          1        0   112
#> [620,]  88   4      0          4        0   100
#> [621,]  88   4      0          4        1   115
#> [622,]  86   9      1          7        1   142
#> [623,]  87   2      0        180        1   130
#> [624,]  88   3      0        115        0   110
#> [625,]  88   2      0        180        1    68
#> [626,]  83   3      0          3        1   130
#> [627,]  87   8      0          8        1   157
#> [628,]  86  15      1          8        1   109
#> [629,]  88   4      0          4        0    86
#> [630,]  89   4      0          4        1   153
#> [631,]  89   5      0        119        1   140
#> [632,]  87   6      0        180        1   110
#> [633,]  87   1      0          1        0   170
#> [634,]  84   8      0        180        1   119
#> [635,]  85   8      0          8        1   136
#> [636,]  84   2      0        110        1   174
#> [637,]  87  15      1          9        1   138
#> [638,]  84   0      0        180        1   136
#> [639,]  89  10      0         46        1   170
#> [640,]  88   1      0          1        0   135
#> [641,]  87   2      0        180        0   160
#> [642,]  87   6      1          0        0   125
#> [643,]  86   3      1          0        1    80
#> [644,]  88   8      0         50        1   154
#> [645,]  90  11      1         10        1   186
#> [646,]  87   6      0        126        1   168
#> [647,]  86  10      0        180        1   137
#> [648,]  90   4      1          0        0   121
#> [649,]  91   1      0          1        1    74
#> [650,]  87  43      0        178        1   130
#> [651,]  88   3      1          2        0   159
#> [652,]  89   3      1          1        1   160
#> [653,]  91   3      0         33        1   137
#> [654,]  91   5      0        169        1   176
#> [655,]  92   7      0          7        1   110
#> [656,]  91   0      0          0        0     0
#> [657,]  89  14      0        180        1    84
#> [658,]  90  18      0        180        0   188
#> [659,]  91   4      1          0        1   120
#> [660,]  90  19      1         11        1   129
#> [661,]  91   2      0          2        1   116
#> [662,]  94   8      0          8        1   142
#> [663,]  91   1      0        180        0   158
#> [664,]  96   3      0         12        1    97
#> [665,]  95   8      1          5        1   150
#> [666,]  94   3      0         26        1   144
#> [667,]  91   7      0          7        0   135
#> [668,]  93   0      1          0        1   122
#> [669,]  93   4      0        180        1   135
#> [670,]  96  15      1          0        1   140
#> 
#> $y
#>   [1] 180.0+   2.0+   5.0+ 180.0+   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [11]   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0+ 180.0+ 180.0+
#>  [21] 180.0+   3.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 180.0+ 180.0+
#>  [31] 180.0+ 180.0+   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [41]   6.0+ 180.0+ 180.0+ 180.0+  73.0  180.0+ 180.0+ 180.0+ 180.0+   5.0+
#>  [51] 180.0+ 180.0+ 177.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  10.0+
#>  [61] 180.0+ 180.0+ 180.0+   7.0  180.0+ 180.0+   1.0  180.0+ 179.0+ 180.0+
#>  [71] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   7.0 
#>  [81]  36.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [91] 180.0+ 180.0+ 180.0+   4.0+ 180.0+ 180.0+ 166.0+ 180.0+  16.0+ 180.0+
#> [101] 152.0+   7.0+   6.0+ 180.0+ 180.0+  13.0+ 171.0+ 180.0+   6.0+   1.0 
#> [111] 180.0+ 180.0+ 175.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [121] 180.0+ 180.0+  16.0+ 180.0+  16.0   15.0+ 180.0+ 134.0+ 180.0+   8.0 
#> [131]   3.0+ 180.0+ 180.0+ 140.0  180.0+   1.0  165.0  180.0+ 180.0+ 180.0+
#> [141] 180.0+   8.0+ 180.0+ 180.0+   0.5  180.0+ 171.0+ 166.0+   5.0+ 180.0+
#> [151]   4.0+ 147.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+   9.0+   1.0  180.0+
#> [161]   2.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0  180.0+ 180.0+ 180.0+ 180.0+
#> [171] 180.0+ 180.0+ 180.0+   9.0+ 180.0+   0.5  180.0+ 161.0+ 171.0+ 180.0+
#> [181]   3.0    1.0  180.0+ 180.0+  10.0+ 180.0+  17.0  180.0+ 180.0+ 180.0+
#> [191] 172.0+  24.0    8.0  180.0+ 180.0+   1.0+ 180.0+  77.0    8.0+  94.0 
#> [201] 180.0+ 169.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0   18.0  180.0+
#> [211] 180.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+ 180.0+   9.0   22.0  180.0+
#> [221]  84.0    7.0+ 180.0+ 180.0+   1.0    1.0  180.0+ 180.0+   4.0+   3.0+
#> [231] 167.0    6.0+ 180.0+ 180.0+ 180.0+  36.0  180.0+   3.0+ 180.0+ 180.0+
#> [241] 180.0+   0.5  180.0+  12.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   9.0 
#> [251] 180.0+ 180.0+ 180.0+  18.0+ 180.0+   3.0+   2.0+ 103.0   15.0  180.0+
#> [261] 180.0+  13.0  179.0+  14.0+   0.5+ 180.0+   3.0+ 180.0+ 175.0+ 180.0+
#> [271]   7.0+   5.0    1.0  180.0+ 123.0+  18.0   11.0+  79.0   80.0   15.0+
#> [281] 180.0+  15.0  180.0+ 174.0+ 180.0+ 180.0+ 180.0+ 180.0+  10.0  180.0+
#> [291] 180.0+ 180.0+   6.0  180.0+  19.0+  12.0   99.0  180.0+ 180.0+ 180.0+
#> [301]  18.0    7.0+ 180.0+ 152.0+ 180.0+  93.0   21.0+   1.0    4.0    5.0 
#> [311]   7.0+ 150.0  180.0+ 180.0+ 180.0+   1.0  180.0+ 180.0+   6.0  180.0+
#> [321]   0.5  180.0+ 180.0+   7.0+   2.0    3.0+ 180.0+ 180.0+  97.0    7.0 
#> [331]   8.0+  18.0  172.0+ 180.0+   7.0    8.0+ 123.0  180.0+ 180.0+  51.0 
#> [341] 180.0+   1.0    1.0   76.0  180.0+ 132.0   10.0+ 180.0+ 162.0  124.0 
#> [351] 180.0+ 180.0+  12.0  180.0+ 180.0+ 180.0+  76.0  180.0+ 180.0+ 180.0+
#> [361]  28.0  180.0+ 180.0+  16.0+ 180.0+  16.0+ 180.0+ 180.0+   6.0  180.0+
#> [371] 180.0+   7.0+  13.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  20.0  180.0+
#> [381]  20.0    0.5  180.0+ 180.0+   3.0   87.0   12.0    4.0+  58.0  180.0+
#> [391] 180.0+ 180.0+ 180.0+   3.0  175.0    2.0  180.0+  14.0+ 180.0+  10.0+
#> [401] 180.0+ 180.0+   8.0+ 179.0+ 159.0   15.0  180.0+  13.0    4.0+ 180.0+
#> [411] 104.0+ 180.0+   3.0+   5.0  180.0+  12.0  180.0+ 180.0+ 180.0+ 180.0+
#> [421] 180.0+  34.0  180.0+ 180.0+   5.0    4.0+ 180.0+   7.0  180.0+   3.0 
#> [431] 180.0+ 180.0+ 180.0+   6.0  180.0+ 180.0+  85.0  180.0+  17.0+ 180.0+
#> [441]   4.0    7.0    0.5  180.0+  12.0  180.0+ 180.0+   4.0    1.0  180.0+
#> [451]  52.0  180.0+ 180.0+  12.0  180.0+ 180.0+ 180.0+ 180.0+  33.0    5.0 
#> [461] 180.0+  12.0    7.0+   3.0  180.0+ 180.0+ 176.0+ 180.0+  47.0  180.0+
#> [471]   7.0  180.0+ 180.0+  10.0  172.0   12.0  180.0+   8.0  180.0+ 180.0+
#> [481]  80.0    4.0+   2.0  152.0+  29.0   24.0   32.0   23.0    6.0  180.0+
#> [491]   3.0+ 180.0+ 180.0+ 180.0+ 180.0+   1.0   11.0    4.0    4.0   10.0+
#> [501] 180.0+   3.0+ 180.0+ 171.0    1.0   43.0  180.0+   6.0  138.0  180.0+
#> [511] 180.0+   8.0   40.0   17.0  161.0  180.0+ 164.0  118.0  173.0  180.0+
#> [521] 175.0+  22.0    3.0  171.0+ 166.0+  20.0+   1.0    3.0+  10.0  180.0+
#> [531]  85.0   10.0    6.0+   6.0  180.0+   5.0    1.0  180.0+ 180.0+ 108.0 
#> [541] 180.0+ 125.0  180.0+   4.0    6.0  180.0+ 180.0+ 103.0  180.0+ 169.0 
#> [551]  70.0  180.0+ 180.0+ 180.0+ 180.0+   7.0+ 180.0+  20.0    8.0+  16.0 
#> [561] 180.0+ 180.0+ 180.0+   0.5  180.0+   2.0  128.0  167.0   62.0  180.0+
#> [571]   4.0   38.0   90.0  180.0+ 180.0+ 180.0+   4.0   71.0   19.0   30.0 
#> [581] 180.0+ 180.0+ 180.0+ 154.0    2.0  180.0+   4.0+ 180.0+  12.0   16.0+
#> [591]   5.0+ 180.0+  83.0   88.0  180.0+   3.0    3.0+ 180.0+ 180.0+ 180.0+
#> [601] 180.0+   4.0+   5.0  180.0+   6.0   70.0  180.0+ 180.0+ 180.0+  14.0 
#> [611]   3.0   13.0   92.0  180.0+  38.0  177.0    6.0+  62.0    6.0+   4.0+
#> [621]   4.0   11.0  180.0+ 115.0  180.0+   3.0+   8.0+ 180.0+   4.0    4.0 
#> [631] 119.0  180.0+   1.0+ 180.0+   8.0  110.0  180.0+ 180.0+  46.0    1.0+
#> [641] 180.0+  25.0    3.0   50.0   11.0  126.0  180.0+   4.0    1.0  178.0+
#> [651]  75.0    3.0+  33.0  169.0    7.0    0.5  180.0+ 180.0+   4.0  180.0+
#> [661]   2.0    8.0+ 180.0+  12.0    8.0   26.0    7.0+   0.5  180.0+  15.0+
#> 
#> $weights
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Warning: Multiple lambdas have been fit. Lambda will be set to 0.01 (see parameter 's').

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8342957