GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::glmnet() from package glmnet.

Details

This learner returns two prediction types:

lp: a vector of linear predictors (relative risk scores), one per observation. Calculated using glmnet::predict.coxnet().

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.coxnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used.

Caution: This learner is different to learners calling glmnet::cv.glmnet() in that it does not use the internal optimization of parameter lambda. Instead, lambda needs to be tuned by the user (e.g., via mlr3tuning). When lambda is tuned, the glmnet will be trained for each tuning iteration. While fitting the whole path of lambdas would be more efficient, as is done by default in glmnet::glmnet(), tuning/selecting the parameter at prediction time (using parameter s) is currently not supported in mlr3 (at least not in efficient manner). Tuning the s parameter is, therefore, currently discouraged.

When the data are i.i.d. and efficiency is key, we recommend using the respective auto-tuning counterpart in mlr_learners_surv.cv_glmnet(). However, in some situations this is not applicable, usually when data are imbalanced or not i.i.d. (longitudinal, time-series) and tuning requires custom resampling strategies (blocked design, stratification).

Custom mlr3 parameters

family is set to "cox" and cannot be changed.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exact	logical	FALSE	TRUE, FALSE	-

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

newoffset	untyped	-		-

nlambda	integer	100		\([1, \infty)\)

offset	untyped	NULL		-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	0.01		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.logistic	character	Newton	Newton, modified.Newton	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvGlmnet

Methods

Public methods

LearnerSurvGlmnet$new()

LearnerSurvGlmnet$selected_features()

LearnerSurvGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.glmnet")
print(learner)
#> <LearnerSurvGlmnet:surv.glmnet>: Regularized Generalized Linear Model
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3proba, mlr3extralearners, glmnet
#> * Predict Types:  [crank], distr, lp
#> * Feature Types: logical, integer, numeric
#> * Properties: selected_features, weights

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#>    Df  %Dev   Lambda
#> 1   0  0.00 0.178700
#> 2   1  0.55 0.162900
#> 3   2  1.50 0.148400
#> 4   3  2.38 0.135200
#> 5   3  5.74 0.123200
#> 6   3  8.14 0.112300
#> 7   3  9.96 0.102300
#> 8   3 11.39 0.093190
#> 9   3 12.55 0.084920
#> 10  3 13.48 0.077370
#> 11  4 14.29 0.070500
#> 12  4 14.97 0.064240
#> 13  4 15.54 0.058530
#> 14  4 16.02 0.053330
#> 15  4 16.42 0.048590
#> 16  5 16.78 0.044270
#> 17  5 17.08 0.040340
#> 18  6 17.34 0.036760
#> 19  6 17.58 0.033490
#> 20  6 17.79 0.030520
#> 21  6 17.96 0.027810
#> 22  6 18.10 0.025340
#> 23  6 18.22 0.023080
#> 24  6 18.33 0.021030
#> 25  6 18.41 0.019170
#> 26  6 18.49 0.017460
#> 27  6 18.55 0.015910
#> 28  6 18.60 0.014500
#> 29  6 18.64 0.013210
#> 30  6 18.68 0.012040
#> 31  6 18.71 0.010970
#> 32  6 18.74 0.009993
#> 33  6 18.76 0.009105
#> 34  6 18.78 0.008296
#> 35  6 18.79 0.007559
#> 36  6 18.81 0.006888
#> 37  6 18.82 0.006276
#> 38  6 18.82 0.005718
#> 39  6 18.83 0.005210
#> 40  6 18.84 0.004747
#> 41  6 18.84 0.004326
#> 42  6 18.85 0.003941
#> 43  6 18.85 0.003591
#> 44  6 18.85 0.003272
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  32   5      1          0        1   121
#>   [2,]  33   2      0          2        0   150
#>   [3,]  35   5      1          2        0   172
#>   [4,]  35  10      1          9        0   106
#>   [5,]  35   2      0        180        0   121
#>   [6,]  37   9      0        180        1   151
#>   [7,]  38  13      1          0        1   161
#>   [8,]  38   2      0        115        0   150
#>   [9,]  36   1      0        180        1   155
#>  [10,]  38  12      1          8        1   120
#>  [11,]  36   5      1          0        1   115
#>  [12,]  33   6      1          1        1   115
#>  [13,]  38  16      1         10        0   160
#>  [14,]  38  12      1         11        1    92
#>  [15,]  40  12      1          9        0   153
#>  [16,]  42   3      1          1        1   130
#>  [17,]  40   2      1          1        1   148
#>  [18,]  38   5      1          3        0   125
#>  [19,]  42   2      0          2        0   140
#>  [20,]  40   6      0        180        1   138
#>  [21,]  42   2      0        180        0   100
#>  [22,]  41   2      1          1        0   166
#>  [23,]  40   1      1          0        1   145
#>  [24,]  43   4      1          0        1   130
#>  [25,]  42   4      0        180        0   162
#>  [26,]  40   3      1          1        0   170
#>  [27,]  44   5      1          1        0   170
#>  [28,]  41  10      1          8        0   150
#>  [29,]  44   3      0        180        0   141
#>  [30,]  45   9      1          7        0   110
#>  [31,]  45   6      0        180        1   170
#>  [32,]  41   5      1          4        1   141
#>  [33,]  43   2      0        180        1   140
#>  [34,]  47   4      1          3        0   118
#>  [35,]  48  15      0        180        1   160
#>  [36,]  44   3      1          0        1   180
#>  [37,]  46   7      1          2        0   166
#>  [38,]  43  29      0        180        1   180
#>  [39,]  45   4      1          0        0   124
#>  [40,]  47   6      1          0        1   116
#>  [41,]  46  13      1         10        0   100
#>  [42,]  47   4      1          3        1   160
#>  [43,]  43   3      1          0        1   124
#>  [44,]  49   5      0         73        1   136
#>  [45,]  45   5      0          5        0   141
#>  [46,]  46   2      1          1        1   122
#>  [47,]  44   4      1          0        1   114
#>  [48,]  44   9      1          8        1   135
#>  [49,]  45   5      0        180        1   190
#>  [50,]  46  15      0        180        1   120
#>  [51,]  45   9      1          0        1   145
#>  [52,]  47   3      1          1        1   120
#>  [53,]  48   3      0        180        0   154
#>  [54,]  48  12      1         11        0   200
#>  [55,]  47   5      1          3        1   130
#>  [56,]  46   3      1          0        1   119
#>  [57,]  50   1      1          0        1   129
#>  [58,]  48   2      1          0        0   184
#>  [59,]  47   7      0        180        0   145
#>  [60,]  50   4      1          1        0   125
#>  [61,]  50   6      1          2        1   140
#>  [62,]  49   7      1          7        1   110
#>  [63,]  46   3      1          1        1   140
#>  [64,]  46   9      1          9        1   122
#>  [65,]  50   7      0        180        1   110
#>  [66,]  49   2      0          2        0   105
#>  [67,]  49  15      1         11        1   160
#>  [68,]  49  23      0        179        1   112
#>  [69,]  46   6      1          0        1   156
#>  [70,]  52   2      0        180        1   170
#>  [71,]  50   7      1          0        1    92
#>  [72,]  49   7      1          4        1    90
#>  [73,]  51   8      0        180        1   140
#>  [74,]  52   2      0        180        0   155
#>  [75,]  46   3      0        180        1   120
#>  [76,]  46   1      1          1        0   145
#>  [77,]  48   7      1          0        1   110
#>  [78,]  53   8      0         36        1   160
#>  [79,]  47   2      1          1        0   110
#>  [80,]  52   4      1          4        0   152
#>  [81,]  49   9      1          3        0   102
#>  [82,]  54  17      1         12        1   102
#>  [83,]  53   5      0         77        0   159
#>  [84,]  53   7      1          0        0   199
#>  [85,]  51   3      1          1        0   140
#>  [86,]  50   2      0          5        1   106
#>  [87,]  50  14      1         13        0   170
#>  [88,]  53   8      1          7        0   160
#>  [89,]  51  25      1          1        0   202
#>  [90,]  49   5      1          2        1   150
#>  [91,]  53   4      0          4        0   140
#>  [92,]  48   6      0        180        0   160
#>  [93,]  48  11      1         10        0   120
#>  [94,]  53   4      1          0        1   156
#>  [95,]  49  16      0         16        0   125
#>  [96,]  55   3      1          1        0   150
#>  [97,]  54  23      1         10        0   131
#>  [98,]  52   7      1          2        0   154
#>  [99,]  54   9      1          1        0   130
#> [100,]  55   4      1          2        0   150
#> [101,]  50   3      0        174        1   153
#> [102,]  55  28      1         13        1   160
#> [103,]  49   1      0          1        1   110
#> [104,]  50   7      1          0        1   127
#> [105,]  56   4      1          1        1   130
#> [106,]  55   2      0          2        0   145
#> [107,]  54   1      0        180        0   162
#> [108,]  54   7      1          0        1   100
#> [109,]  56   3      0        180        1   193
#> [110,]  56   2      0        180        0   132
#> [111,]  52   8      0        180        0   119
#> [112,]  53  18      1          9        1   150
#> [113,]  54   3      0        180        1   180
#> [114,]  55   6      0        180        0   170
#> [115,]  52  16      0         16        0   152
#> [116,]  53  10      1          9        0   172
#> [117,]  52  16      1         14        0   170
#> [118,]  53   4      0        180        1   150
#> [119,]  54  12      1          0        1   190
#> [120,]  54   3      0        180        0   128
#> [121,]  55   1      0          2        0   130
#> [122,]  57   3      0          3        0   120
#> [123,]  54   2      1          1        0   135
#> [124,]  57   5      1          3        1   138
#> [125,]  57   1      0        180        1   156
#> [126,]  57   1      0          1        1   100
#> [127,]  52  15      1         14        0   130
#> [128,]  56  14      1         11        0   130
#> [129,]  53   3      1          0        1   200
#> [130,]  57  10      0        180        1   170
#> [131,]  55   3      1          1        1   156
#> [132,]  53  21      1         13        1   130
#> [133,]  57   4      0        180        1   119
#> [134,]  55   9      1          2        1   147
#> [135,]  55  13      0        166        1   140
#> [136,]  56   5      0          5        1   150
#> [137,]  54  23      1          8        0   120
#> [138,]  57   4      1          2        1   185
#> [139,]  57  11      1         10        1   129
#> [140,]  55   3      1          2        0   140
#> [141,]  54   7      1          0        1   141
#> [142,]  56   4      0          4        0   164
#> [143,]  59  15      1         10        0   140
#> [144,]  58   9      1          0        1   180
#> [145,]  55   5      1          0        0   140
#> [146,]  59   9      1          1        1   125
#> [147,]  60  11      1          9        0   106
#> [148,]  59   3      0        180        0   120
#> [149,]  57   2      0          2        1   120
#> [150,]  60   5      1          1        0   138
#> [151,]  57   5      0        180        1   130
#> [152,]  57  10      1          9        0   103
#> [153,]  59   5      0        180        1   155
#> [154,]  59   4      1          0        1   152
#> [155,]  61   9      0          9        1   160
#> [156,]  58   4      1          3        0   120
#> [157,]  60   0      1          0        1    80
#> [158,]  59   2      1          1        0   140
#> [159,]  61   4      1          3        0   151
#> [160,]  61   9      1          8        0   150
#> [161,]  61   3      1          2        1   102
#> [162,]  61  20      1         13        0   130
#> [163,]  58  10      0         10        1   150
#> [164,]  57   4      1          3        0   138
#> [165,]  61   3      0         17        0   143
#> [166,]  56  14      0         45        0   130
#> [167,]  58  19      1         13        1   140
#> [168,]  56  13      1          6        1   158
#> [169,]  56  18      1         11        1   165
#> [170,]  59   9      1          0        1    80
#> [171,]  55   4      1          3        1   160
#> [172,]  58  11      0        172        1   135
#> [173,]  55   9      1          7        1   135
#> [174,]  61   4      1          0        1   115
#> [175,]  56   8      1          8        0   120
#> [176,]  59  11      1          8        1   190
#> [177,]  57   1      0          1        0   126
#> [178,]  57  15      1         13        1   110
#> [179,]  58   5      1          1        1   135
#> [180,]  61   8      0         77        0   120
#> [181,]  61  13      0         13        0   210
#> [182,]  58   8      1          5        0   152
#> [183,]  62  10      1          0        1   153
#> [184,]  62   7      1          2        1   180
#> [185,]  61  18      0        170        0   140
#> [186,]  61  28      1          7        0   133
#> [187,]  57   7      0        169        0   180
#> [188,]  59  13      1          2        0   198
#> [189,]  57  12      1          9        1   120
#> [190,]  62   4      1          0        0   160
#> [191,]  58   3      1          0        1   146
#> [192,]  62   4      1          3        0   173
#> [193,]  58   2      0         30        0   202
#> [194,]  59  16      1          9        1   133
#> [195,]  63   6      0         28        1   120
#> [196,]  61   5      0          5        1   110
#> [197,]  61   5      0          5        1   160
#> [198,]  58  11      1          9        0   179
#> [199,]  57   2      1          1        0   159
#> [200,]  62  17      1         10        1   180
#> [201,]  62   1      1          0        1   172
#> [202,]  58   7      0        180        1   150
#> [203,]  61   7      0        180        0   135
#> [204,]  63   4      1          3        0   222
#> [205,]  63   4      0        180        1   190
#> [206,]  63  15      1         10        1   126
#> [207,]  64   4      0        180        0   130
#> [208,]  60  18      1         13        0   132
#> [209,]  59   8      0        180        1   140
#> [210,]  59   1      0         22        1   162
#> [211,]  60   7      1          5        1   141
#> [212,]  60   7      0          7        0   140
#> [213,]  60   7      1          1        1    90
#> [214,]  65  13      0        180        1   100
#> [215,]  63   1      0          1        0   162
#> [216,]  63   1      0          1        0   130
#> [217,]  62   6      0        180        0   170
#> [218,]  59   4      0          4        0   149
#> [219,]  60   3      0          3        0   168
#> [220,]  64  10      1          9        0   160
#> [221,]  60   8      0         17        1   130
#> [222,]  64  12      1         11        0   160
#> [223,]  66   1      1          0        1   120
#> [224,]  64   6      1          0        1   140
#> [225,]  63  10      1          0        1   148
#> [226,]  63  14      1          9        0   123
#> [227,]  65  36      1         11        0   140
#> [228,]  63   4      1          3        0   162
#> [229,]  66   3      1          1        0   127
#> [230,]  64  32      1          9        1   160
#> [231,]  65  10      1          8        1   120
#> [232,]  64  21      1         10        0   190
#> [233,]  61  12      1         11        0   154
#> [234,]  61   4      0        180        1   113
#> [235,]  63  16      1          7        1   110
#> [236,]  66   6      1          1        1   130
#> [237,]  62   3      1          1        1   199
#> [238,]  65   6      0          9        0   112
#> [239,]  63   5      1          4        0   170
#> [240,]  63   2      1          1        0   180
#> [241,]  62  13      1         11        0   180
#> [242,]  67  11      0         11        1   100
#> [243,]  66  18      1          5        0   142
#> [244,]  66  16      1         11        1   169
#> [245,]  62   9      0        180        0   145
#> [246,]  61  14      1          5        0   140
#> [247,]  61  15      1         10        0   130
#> [248,]  63   9      1          8        1   160
#> [249,]  63   3      1          2        0   120
#> [250,]  63   2      1          0        0   140
#> [251,]  65   8      1          0        1   140
#> [252,]  67   6      0        180        1   170
#> [253,]  65  15      1         11        1   160
#> [254,]  68   5      1          4        1   150
#> [255,]  64  13      1         12        1   150
#> [256,]  64   6      1          0        1   125
#> [257,]  66   7      1          0        1   115
#> [258,]  64  14      1         13        1   150
#> [259,]  64   0      0          0        1   148
#> [260,]  67   4      1          3        0   130
#> [261,]  65   2      1          1        1   170
#> [262,]  64  10      1          9        1   110
#> [263,]  67   8      1          1        1   130
#> [264,]  68   5      0          5        1    90
#> [265,]  63  10      0         16        1   160
#> [266,]  66  14      0        180        0   130
#> [267,]  64   1      0          1        1   120
#> [268,]  68  18      0        180        1   260
#> [269,]  63   1      1          0        1   155
#> [270,]  63  10      0         18        1   130
#> [271,]  67  11      0         11        0   150
#> [272,]  66  12      1         10        1   150
#> [273,]  65  15      1         12        1   150
#> [274,]  66  11      1          0        0   100
#> [275,]  66  15      1         13        1   160
#> [276,]  65  11      1          6        0   130
#> [277,]  69   6      0        180        1   100
#> [278,]  66   9      1          8        0   130
#> [279,]  65   8      1          0        1    90
#> [280,]  66   3      0          3        1   138
#> [281,]  69   1      1          0        0   170
#> [282,]  67   1      0        180        1   160
#> [283,]  65   1      1          0        0   133
#> [284,]  67   7      1          4        1   130
#> [285,]  63   2      1          0        0    99
#> [286,]  67   2      0        180        0   184
#> [287,]  65   6      0          6        0    80
#> [288,]  65  10      1          1        1   148
#> [289,]  66  19      1         12        1   150
#> [290,]  67  12      1         12        0   160
#> [291,]  69   6      0         99        1   140
#> [292,]  65   4      1          1        0   130
#> [293,]  64   4      0        179        0   160
#> [294,]  66   4      0        180        1   130
#> [295,]  64   4      0        180        1   140
#> [296,]  64   0      1          0        1   118
#> [297,]  67   2      0         18        0   131
#> [298,]  66   7      1          5        1   131
#> [299,]  69   4      1          3        1   150
#> [300,]  65  13      1         12        1   130
#> [301,]  69  17      1         10        0   140
#> [302,]  69   8      0         93        0   140
#> [303,]  64  21      0         21        1   155
#> [304,]  66   6      0        180        0   140
#> [305,]  65   1      0          1        1   120
#> [306,]  68  18      1          0        1   160
#> [307,]  68   4      0          4        1   190
#> [308,]  71   3      0          5        0   112
#> [309,]  68   7      0        150        0   210
#> [310,]  71  20      1          0        1   160
#> [311,]  67   2      0        180        0   128
#> [312,]  66   9      1          3        1   151
#> [313,]  70   4      1          0        1   180
#> [314,]  69   8      0        180        1   153
#> [315,]  67  10      1          9        0   200
#> [316,]  69   0      0          0        1   148
#> [317,]  65   2      0        180        0   130
#> [318,]  69   3      1          2        0   151
#> [319,]  67  14      1         13        0   130
#> [320,]  65  14      1         13        1   150
#> [321,]  69   8      0        180        1   180
#> [322,]  66   2      0          2        1   228
#> [323,]  69   5      0          5        1   142
#> [324,]  69   3      0          3        1   130
#> [325,]  70  22      1         13        0   103
#> [326,]  67   5      0          5        0   130
#> [327,]  68   6      0        180        0   145
#> [328,]  69   8      1          5        1   195
#> [329,]  72   3      1          0        1   132
#> [330,]  72   7      0          7        1   110
#> [331,]  69   8      1          7        1   108
#> [332,]  67   3      0        180        0   110
#> [333,]  66   2      1          1        0   123
#> [334,]  69  19      0        180        0   130
#> [335,]  68  18      0         18        1   100
#> [336,]  69  11      1          0        1   120
#> [337,]  66   2      0        180        0   130
#> [338,]  67   7      1          4        0   122
#> [339,]  69   4      1          3        0   132
#> [340,]  68   2      0          7        1   130
#> [341,]  69   8      1          2        0   121
#> [342,]  67  13      1          9        0   130
#> [343,]  70   3      0        123        0   130
#> [344,]  70   9      0        180        1   142
#> [345,]  72   5      1          4        0   170
#> [346,]  67  22      1          1        1   140
#> [347,]  68   3      0         19        0   135
#> [348,]  67  12      1          8        0   120
#> [349,]  69   1      0          1        1   110
#> [350,]  67   4      0         60        1   136
#> [351,]  69   5      0         76        0   120
#> [352,]  67   8      1          0        1   130
#> [353,]  70   7      0          7        0   102
#> [354,]  68   7      1          2        0   135
#> [355,]  73  20      1          0        1   170
#> [356,]  71   6      0          9        0   120
#> [357,]  69  10      1          6        1   120
#> [358,]  70  11      0        180        1   210
#> [359,]  72  19      1          8        0   120
#> [360,]  67   8      0        180        1   170
#> [361,]  67   5      1          0        1   147
#> [362,]  67   9      0        180        0   158
#> [363,]  73  13      0        152        1   130
#> [364,]  70   5      0        180        0   150
#> [365,]  67   4      1          1        0   134
#> [366,]  72   6      1          5        0   115
#> [367,]  71   1      0        173        1   188
#> [368,]  68  23      0        180        1   220
#> [369,]  70   3      0        180        0   121
#> [370,]  69   3      0        180        0   220
#> [371,]  68   4      1          3        0   210
#> [372,]  72   5      0         28        0   120
#> [373,]  73   6      0        180        1   117
#> [374,]  69  16      1         10        1   140
#> [375,]  69   8      1          1        0   164
#> [376,]  68   7      0        180        1   130
#> [377,]  69   1      1          0        0   155
#> [378,]  71   2      1          0        1   180
#> [379,]  73   7      0          7        1   140
#> [380,]  72   6      0        180        1   130
#> [381,]  74   8      1          0        1    85
#> [382,]  73   4      0        180        1   154
#> [383,]  71   3      1          1        0   150
#> [384,]  71  15      1         11        0   165
#> [385,]  74  20      0         20        1   180
#> [386,]  68   9      0        180        1   120
#> [387,]  71  20      1         10        0   140
#> [388,]  73   3      1          0        1   136
#> [389,]  70   5      1          0        1   190
#> [390,]  71  17      1         11        0   160
#> [391,]  71   3      1          2        1   190
#> [392,]  73  10      1          8        0   106
#> [393,]  69  12      1          1        1   149
#> [394,]  70  26      1         11        1   120
#> [395,]  74   4      0          4        0   120
#> [396,]  73   4      0         58        1   160
#> [397,]  72   5      1          3        1   160
#> [398,]  70   3      0        180        1   154
#> [399,]  73   6      0        180        0   110
#> [400,]  72  15      1          0        1   150
#> [401,]  71   7      1          2        0   143
#> [402,]  74   3      0          3        1   150
#> [403,]  71  13      1          8        0   121
#> [404,]  71  14      1         13        1   170
#> [405,]  72  10      1          8        1   153
#> [406,]  69   7      0        180        1   144
#> [407,]  72  15      1         13        0   156
#> [408,]  70   8      0          8        0   120
#> [409,]  71  10      1          9        1   120
#> [410,]  75   2      1          1        0   145
#> [411,]  73  10      1          9        1   146
#> [412,]  72  10      1          9        1   160
#> [413,]  73  10      1         10        1   120
#> [414,]  73   1      0          1        1    80
#> [415,]  75   9      1          7        0   140
#> [416,]  71  11      1          8        0   110
#> [417,]  71   4      0          4        0   134
#> [418,]  72  15      1         12        1   120
#> [419,]  73  10      1          8        0   120
#> [420,]  70   7      1          4        0   184
#> [421,]  72   1      1          1        0   168
#> [422,]  72   7      0         57        1   145
#> [423,]  73  10      0        180        0   162
#> [424,]  72  11      0         11        1   140
#> [425,]  73   5      1          3        1   112
#> [426,]  75   1      0        180        1   140
#> [427,]  71   3      1          0        0   144
#> [428,]  73   5      0        180        0   126
#> [429,]  73   4      0        180        0   124
#> [430,]  76   3      1          0        1   120
#> [431,]  71  32      1         12        1   107
#> [432,]  72   5      0        180        0   154
#> [433,]  72   3      0        180        0   160
#> [434,]  76   5      0          5        1   130
#> [435,]  77  11      0         11        1   150
#> [436,]  72   7      1          2        0   142
#> [437,]  71  16      0        180        0   140
#> [438,]  73  10      1         10        0   124
#> [439,]  74   7      0        180        1   150
#> [440,]  74   2      1          1        0   140
#> [441,]  74  19      1          4        1   200
#> [442,]  75  23      1         14        1   110
#> [443,]  74   2      0        180        0   190
#> [444,]  72   4      0         85        1   120
#> [445,]  76  13      1         10        0   110
#> [446,]  75   0      0          0        1   130
#> [447,]  73  13      1         11        0   195
#> [448,]  75  12      0         12        1   160
#> [449,]  76  13      1          8        1   148
#> [450,]  75   4      1          2        1   188
#> [451,]  74   6      0        180        0   160
#> [452,]  76   4      0          4        1   155
#> [453,]  75   1      0          1        1   125
#> [454,]  73   1      0         52        1   105
#> [455,]  73   0      0        180        0   156
#> [456,]  72   5      0        180        0   120
#> [457,]  78  12      1         11        1   160
#> [458,]  76   5      0        180        0   185
#> [459,]  74  10      1          0        1   135
#> [460,]  76   5      1          0        1   167
#> [461,]  74   8      1          8        1   170
#> [462,]  77   5      1          0        0   123
#> [463,]  77  12      1          9        1   100
#> [464,]  73  10      1          9        0   146
#> [465,]  77  12      0        180        0   130
#> [466,]  77   1      1          0        1    90
#> [467,]  76  12      1         11        1   120
#> [468,]  78   5      1          0        1   170
#> [469,]  75   3      1          1        1   171
#> [470,]  74   9      1          8        0   200
#> [471,]  75   6      0        180        0   150
#> [472,]  77   3      0        180        0   110
#> [473,]  76  29      0         47        0    90
#> [474,]  73  11      1          2        1   110
#> [475,]  74  15      0        180        1   172
#> [476,]  78   8      1          6        1   110
#> [477,]  74   7      0          7        0   161
#> [478,]  76  13      1          1        1   170
#> [479,]  78  32      1          9        1   198
#> [480,]  78   0      0        180        1   212
#> [481,]  78  13      1          5        0   130
#> [482,]  75   5      0        119        1   150
#> [483,]  75  12      1          1        1   120
#> [484,]  78  15      0        180        1   270
#> [485,]  75  13      1          6        0   150
#> [486,]  74  10      1          8        0   135
#> [487,]  78   2      1          1        0   130
#> [488,]  77   2      1          0        1   143
#> [489,]  78  10      0        180        1   130
#> [490,]  76   7      0         29        1   150
#> [491,]  77  24      0         24        1   160
#> [492,]  79   8      0         32        1   120
#> [493,]  80   9      0         23        1   128
#> [494,]  78  11      1          1        1   140
#> [495,]  79  11      0        180        0   160
#> [496,]  79   2      1          0        1   121
#> [497,]  78  14      1          0        1   140
#> [498,]  81   1      0          1        0   130
#> [499,]  78  11      1          8        1   118
#> [500,]  79   4      0          4        1   125
#> [501,]  76  10      1          8        0   180
#> [502,]  77   6      0          6        1   107
#> [503,]  80   3      1          0        1   120
#> [504,]  75   2      1          1        1   204
#> [505,]  76   1      0          1        1   140
#> [506,]  78   7      1          0        1   110
#> [507,]  79   3      0          3        0   120
#> [508,]  77   7      0        180        1   170
#> [509,]  77   6      0          6        1   144
#> [510,]  79   4      1          0        1   120
#> [511,]  81   1      0        180        0   120
#> [512,]  80  15      1         12        1   150
#> [513,]  77   9      1          4        0   141
#> [514,]  82   5      0          8        1   120
#> [515,]  80  40      1          0        1   138
#> [516,]  78   4      0         59        1   112
#> [517,]  80  17      1         12        0   100
#> [518,]  79  10      0         10        1   120
#> [519,]  80   9      0        118        1   186
#> [520,]  80   6      0        173        1   160
#> [521,]  79   1      0         37        1   140
#> [522,]  81   3      0        180        0   184
#> [523,]  81   2      0        175        0   172
#> [524,]  77  13      1          0        1   190
#> [525,]  78  15      0         15        0   165
#> [526,]  80   5      1          1        1   108
#> [527,]  78   4      0        180        0   175
#> [528,]  79   3      0          3        1   101
#> [529,]  78  26      1          5        0   194
#> [530,]  76   1      0        166        0   131
#> [531,]  78  20      1          0        1   109
#> [532,]  77  10      1          8        1   130
#> [533,]  82   3      1          1        1   144
#> [534,]  77   5      0         85        0   188
#> [535,]  79   6      0          6        0   152
#> [536,]  80   6      1          0        1   119
#> [537,]  78   2      0        180        0   148
#> [538,]  80   5      0          5        1   130
#> [539,]  82   1      1          0        1    82
#> [540,]  81   1      0        108        0   129
#> [541,]  78  12      0        180        0   134
#> [542,]  79   1      0        125        0   193
#> [543,]  82  21      1          2        0   155
#> [544,]  80   6      0          6        1   110
#> [545,]  83   9      1          5        1   170
#> [546,]  83   5      0        180        0   148
#> [547,]  79   7      1          6        0   130
#> [548,]  83   4      0        103        0    97
#> [549,]  81  11      1          8        0   160
#> [550,]  81   5      0        177        0    41
#> [551,]  80  11      1          8        0   170
#> [552,]  78  23      1         10        1   145
#> [553,]  79   4      0          4        1   183
#> [554,]  78   9      1          4        1   120
#> [555,]  82   8      1          1        0   128
#> [556,]  81  15      0        180        1   140
#> [557,]  80   7      1          0        1   146
#> [558,]  81  20      1          9        0   170
#> [559,]  83   8      0          8        0   115
#> [560,]  81  16      0         16        1   110
#> [561,]  80   6      1          0        1   150
#> [562,]  81   8      0        180        0   146
#> [563,]  81   2      1          1        0   198
#> [564,]  83   2      0          2        1   155
#> [565,]  82   6      0        128        1   100
#> [566,]  84   4      0        167        0   198
#> [567,]  80   3      1          1        1   230
#> [568,]  82  23      1          0        0   110
#> [569,]  84   5      0        180        1   203
#> [570,]  84   4      0          4        1    85
#> [571,]  81   1      0          1        1   150
#> [572,]  83   3      0        180        0   174
#> [573,]  81   4      0         90        1   138
#> [574,]  79   9      1          8        0   150
#> [575,]  85   3      1          2        1   160
#> [576,]  80   2      1          0        1   130
#> [577,]  80   6      0         71        1   189
#> [578,]  83   1      0          1        1   100
#> [579,]  82  19      0         19        0   120
#> [580,]  83   9      0        180        0   198
#> [581,]  83   3      0        114        0    98
#> [582,]  81  14      1         12        1   128
#> [583,]  83   2      0        154        0   130
#> [584,]  85   9      1          6        1   160
#> [585,]  83   1      0        180        0   160
#> [586,]  81   4      0          4        0   175
#> [587,]  81   1      0          1        1   145
#> [588,]  82  16      1          8        0   103
#> [589,]  82   5      1          0        1   146
#> [590,]  86  12      0        180        1   120
#> [591,]  83  12      1          2        1   170
#> [592,]  81  19      1         14        0   120
#> [593,]  82   3      1          2        0   130
#> [594,]  84   6      0        165        0   145
#> [595,]  86   3      0          3        1   140
#> [596,]  84   3      0        180        1   120
#> [597,]  81  13      0        180        0   152
#> [598,]  85   3      0          3        1   118
#> [599,]  81   2      1          0        1   118
#> [600,]  81   4      0        180        0   160
#> [601,]  83   9      0        180        1   149
#> [602,]  82   1      0        180        1   193
#> [603,]  83   4      0          4        0   130
#> [604,]  86  12      1          0        1   132
#> [605,]  82  14      1         11        1   103
#> [606,]  86   6      1          0        1   140
#> [607,]  83  19      0         43        0   150
#> [608,]  84   3      1          2        0   125
#> [609,]  86   2      0        180        1   169
#> [610,]  83  13      1         12        0   170
#> [611,]  84   7      1          2        0   148
#> [612,]  84   9      0         92        1   110
#> [613,]  84   3      0        180        1   170
#> [614,]  86   4      0         38        1   122
#> [615,]  82   4      0          4        0   130
#> [616,]  85   3      0          3        1   113
#> [617,]  86   6      0          6        1   117
#> [618,]  84  13      0         62        1   100
#> [619,]  86   6      1          1        0   112
#> [620,]  83  20      1          3        1   150
#> [621,]  85  22      0         22        1   184
#> [622,]  83   9      0         65        1   150
#> [623,]  86   9      1          7        1   142
#> [624,]  88   2      0        180        1    68
#> [625,]  83   3      0          3        1   130
#> [626,]  88   4      0          4        0    86
#> [627,]  89   4      0          4        1   153
#> [628,]  89   5      0        119        1   140
#> [629,]  87   1      0          1        0   170
#> [630,]  84   8      0        180        1   119
#> [631,]  85   8      0          8        1   136
#> [632,]  84   0      0        180        1   136
#> [633,]  89  10      0         46        1   170
#> [634,]  90  14      0         14        1   100
#> [635,]  88   1      0          1        0   135
#> [636,]  86   4      0        180        1   145
#> [637,]  91   8      0          8        0   100
#> [638,]  87   2      0        180        0   160
#> [639,]  87   6      1          0        0   125
#> [640,]  86   3      1          0        1    80
#> [641,]  88   7      0         24        0   119
#> [642,]  88   8      0         50        1   154
#> [643,]  90  11      1         10        1   186
#> [644,]  87   6      0        126        1   168
#> [645,]  90   4      1          0        0   121
#> [646,]  91   1      0          1        1    74
#> [647,]  87  43      0        178        1   130
#> [648,]  87   5      0         36        1   150
#> [649,]  89   3      1          1        1   160
#> [650,]  91   3      0         33        1   137
#> [651,]  88   5      0        158        0   100
#> [652,]  87   7      0         74        1   105
#> [653,]  91   5      0        169        1   176
#> [654,]  89  52      0         52        1   130
#> [655,]  92   7      0          7        1   110
#> [656,]  89   4      0          4        1   159
#> [657,]  91   0      0          0        0     0
#> [658,]  89  14      0        180        1    84
#> [659,]  90  19      1         11        1   129
#> [660,]  90   1      0          1        1   118
#> [661,]  91   2      0          2        1   116
#> [662,]  93   8      0        179        1   110
#> [663,]  91   1      0        180        0   158
#> [664,]  90   3      0         67        0   162
#> [665,]  96   3      0         12        1    97
#> [666,]  95   8      1          5        1   150
#> [667,]  91  12      0         53        1   212
#> [668,]  93   4      0        180        1   135
#> [669,]  96   3      1          0        1   104
#> [670,]  96  15      1          0        1   140
#> 
#> $y
#>   [1]   5.0+   2.0+   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 115.0  180.0+  12.0 
#>  [11]   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+   5.0+   2.0+ 180.0+
#>  [21] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 155.0+ 180.0+ 180.0+ 180.0+
#>  [31] 180.0+   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   6.0+
#>  [41] 180.0+ 180.0+ 180.0+  73.0    5.0+ 161.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [51] 177.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 172.0+ 180.0+ 180.0+ 180.0+
#>  [61] 180.0+   7.0  180.0+ 180.0+ 180.0+   2.0  179.0+ 179.0+ 180.0+ 180.0+
#>  [71] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   7.0   36.0  180.0+   4.0+
#>  [81] 180.0+ 180.0+  77.0  180.0+ 180.0+   5.0  180.0+ 180.0+ 180.0+ 180.0+
#>  [91]   4.0+ 180.0+ 180.0+ 166.0+  16.0+ 180.0+ 152.0+   7.0+ 180.0+ 180.0+
#> [101] 174.0+  28.0    1.0  180.0+ 180.0+   2.0  180.0+   7.0+ 180.0+ 180.0+
#> [111] 180.0+ 180.0+ 180.0+ 180.0+  16.0+ 180.0+  16.0  180.0+  12.0+ 180.0+
#> [121]   2.0    3.0+ 180.0+ 140.0  180.0+   1.0  180.0+ 180.0+ 180.0+ 180.0+
#> [131] 180.0+ 180.0+ 180.0+  15.0  166.0+   5.0+ 180.0+   4.0+ 180.0+ 180.0+
#> [141] 180.0+   4.0+ 180.0+   9.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0  180.0+
#> [151] 180.0+ 180.0+ 180.0+ 180.0+   9.0+ 180.0+   0.5  180.0+ 180.0+ 180.0+
#> [161]   3.0  180.0+  10.0+ 180.0+  17.0   45.0   19.0  180.0+ 180.0+   9.0+
#> [171] 180.0+ 172.0+  24.0  180.0+   8.0  180.0+   1.0+  15.0  180.0+  77.0 
#> [181]  13.0+   8.0+ 180.0+ 180.0+ 170.0   94.0  169.0  180.0+ 180.0+ 180.0+
#> [191]   3.0+ 180.0+  30.0  180.0+  28.0    5.0    5.0+ 180.0+ 180.0+ 180.0+
#> [201]   1.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  22.0 
#> [211]  84.0    7.0+ 180.0+ 180.0+   1.0    1.0  180.0+   4.0+   3.0+ 167.0 
#> [221]  17.0   12.0  180.0+ 180.0+ 180.0+  14.0+  36.0  180.0+   3.0+ 180.0+
#> [231] 180.0+ 180.0+  12.0+ 180.0+ 180.0+ 180.0+ 180.0+   9.0  180.0+ 180.0+
#> [241] 180.0+  11.0+  18.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   3.0+   2.0+
#> [251]  15.0  180.0+ 180.0+   5.0+  13.0  180.0+ 179.0+  14.0+   0.5+ 180.0+
#> [261] 175.0+ 180.0+   8.0    5.0   16.0  180.0+   1.0  180.0+   1.0+  18.0 
#> [271]  11.0+  80.0   15.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   8.0+   3.0 
#> [281] 175.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+   6.0  180.0+  19.0+  12.0 
#> [291]  99.0  180.0+ 179.0+ 180.0+ 180.0+   0.5   18.0    7.0+ 152.0+ 180.0+
#> [301] 180.0+  93.0   21.0+ 180.0+   1.0   18.0+   4.0    5.0  150.0  180.0+
#> [311] 180.0+ 180.0+ 180.0+ 180.0+ 174.0+   0.5  180.0+ 180.0+ 180.0+  14.0+
#> [321] 180.0+   2.0    5.0+   3.0+ 180.0+   5.0+ 180.0+ 180.0+ 180.0+   7.0 
#> [331]   8.0+ 180.0+   2.0+ 180.0+  18.0  180.0+ 180.0+   7.0  180.0+   7.0 
#> [341]   8.0+  13.0+ 123.0  180.0+ 180.0+  51.0   19.0  180.0+   1.0   60.0 
#> [351]  76.0  180.0+   7.0+   7.0+ 124.0    9.0  180.0+ 180.0+ 180.0+ 180.0+
#> [361] 180.0+ 180.0+ 152.0  180.0+  76.0  180.0+ 173.0+ 180.0+ 180.0+ 180.0+
#> [371] 180.0+  28.0  180.0+  16.0+ 180.0+ 180.0+ 180.0+ 180.0+   7.0+ 180.0+
#> [381] 180.0+ 180.0+   3.0+ 180.0+  20.0  180.0+  20.0  180.0+ 180.0+ 180.0+
#> [391]   3.0   87.0   12.0  180.0+   4.0+  58.0  180.0+ 180.0+ 180.0+ 180.0+
#> [401] 180.0+   3.0  175.0   14.0+  10.0+ 180.0+ 180.0+   8.0+ 179.0+ 180.0+
#> [411] 180.0+ 159.0   15.0    1.0  180.0+ 180.0+   4.0+ 180.0+  10.0  104.0+
#> [421]   1.0   57.0  180.0+  11.0    5.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [431] 177.0+ 180.0+ 180.0+   5.0   11.0+   7.0  180.0+  10.0  180.0+ 180.0+
#> [441] 180.0+ 180.0+ 180.0+  85.0  174.0+   0.5  180.0+  12.0  180.0+  46.0 
#> [451] 180.0+   4.0    1.0   52.0  180.0+ 180.0+  12.0  180.0+ 180.0+ 180.0+
#> [461]   8.0    5.0  180.0+ 180.0+ 180.0+   1.0   12.0  180.0+   3.0  168.0+
#> [471] 180.0+ 180.0+  47.0   11.0  180.0+   8.0+   7.0  180.0+  32.0  180.0+
#> [481] 172.0  119.0   12.0  180.0+ 180.0+ 180.0+ 180.0+   2.0  180.0+  29.0 
#> [491]  24.0   32.0   23.0  180.0+ 180.0+ 180.0+ 180.0+   1.0   11.0    4.0 
#> [501]  10.0+   6.0    3.0+   2.0+   1.0   43.0    3.0  180.0+   6.0  138.0 
#> [511] 180.0+ 180.0+  71.0    8.0   40.0   59.0   17.0   10.0+ 118.0  173.0 
#> [521]  37.0  180.0+ 175.0+  22.0   15.0+   5.0+ 180.0+   3.0  171.0+ 166.0+
#> [531]  20.0+  10.0  180.0+  85.0    6.0+   6.0  180.0+   5.0    1.0  108.0 
#> [541] 180.0+ 125.0  180.0+   6.0    9.0+ 180.0+ 180.0+ 103.0  180.0+ 177.0+
#> [551] 169.0   70.0    4.0  180.0+ 180.0+ 180.0+   7.0+  20.0    8.0+  16.0 
#> [561] 180.0+ 180.0+ 180.0+   2.0  128.0  167.0    3.0+  62.0  180.0+   4.0 
#> [571]   1.0  180.0+  90.0  180.0+ 180.0+ 180.0+  71.0    1.0   19.0  180.0+
#> [581] 114.0  180.0+ 154.0  180.0+ 180.0+   4.0+   1.0   16.0+   5.0+ 180.0+
#> [591]  77.0  180.0+   3.0  165.0    3.0  180.0+ 180.0+   3.0+ 180.0+ 180.0+
#> [601] 180.0+ 180.0+   4.0+ 180.0+ 174.0    6.0   43.0  180.0+ 180.0+  13.0 
#> [611] 180.0+  92.0  180.0+  38.0    4.0    3.0+   6.0+  62.0    6.0+  20.0 
#> [621]  22.0   65.0   11.0  180.0+   3.0+   4.0    4.0  119.0    1.0+ 180.0+
#> [631]   8.0  180.0+  46.0   14.0    1.0+ 180.0+   8.0  180.0+  25.0    3.0 
#> [641]  24.0   50.0   11.0  126.0    4.0    1.0  178.0+  36.0    3.0+  33.0 
#> [651] 158.0   74.0  169.0   52.0    7.0    4.0    0.5  180.0+ 180.0+   1.0+
#> [661]   2.0  179.0+ 180.0+  67.0   12.0    8.0   53.0  180.0+   3.0   15.0+
#> 
#> $weights
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Warning: Multiple lambdas have been fit. Lambda will be set to 0.01 (see parameter 's').

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8296689