GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.coxnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.coxnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Caution: This learner is different to learners calling glmnet::cv.glmnet() in that it does not use the internal optimization of parameter lambda. Instead, lambda needs to be tuned by the user (e.g., via mlr3tuning). When lambda is tuned, the glmnet will be trained for each tuning iteration. While fitting the whole path of lambdas would be more efficient, as is done by default in glmnet::glmnet(), tuning/selecting the parameter at prediction time (using parameter s) is currently not supported in mlr3 (at least not in efficient manner). Tuning the s parameter is, therefore, currently discouraged.

When the data are i.i.d. and efficiency is key, we recommend using the respective auto-tuning counterpart in mlr_learners_surv.cv_glmnet(). However, in some situations this is not applicable, usually when data are imbalanced or not i.i.d. (longitudinal, time-series) and tuning requires custom resampling strategies (blocked design, stratification).

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.coxnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exact	logical	FALSE	TRUE, FALSE	-

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	TRUE	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	0.01		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.logistic	character	Newton	Newton, modified.Newton	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvGlmnet

Methods

Public methods

LearnerSurvGlmnet$new()

LearnerSurvGlmnet$selected_features()

LearnerSurvGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.glmnet")
print(learner)
#> <LearnerSurvGlmnet:surv.glmnet>: Regularized Generalized Linear Model
#> * Model: -
#> * Parameters: use_pred_offset=TRUE
#> * Packages: mlr3, mlr3proba, mlr3extralearners, glmnet
#> * Predict Types:  [crank], distr, lp
#> * Feature Types: logical, integer, numeric
#> * Properties: offset, selected_features, weights

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#>    Df  %Dev   Lambda
#> 1   0  0.00 0.203000
#> 2   1  0.70 0.185000
#> 3   1  1.28 0.168600
#> 4   1  1.77 0.153600
#> 5   2  2.48 0.139900
#> 6   2  3.14 0.127500
#> 7   3  5.85 0.116200
#> 8   3  7.96 0.105900
#> 9   3  9.55 0.096460
#> 10  3 10.80 0.087890
#> 11  3 11.80 0.080080
#> 12  3 12.61 0.072970
#> 13  3 13.27 0.066480
#> 14  3 13.83 0.060580
#> 15  3 14.29 0.055200
#> 16  3 14.68 0.050290
#> 17  3 15.01 0.045820
#> 18  4 15.29 0.041750
#> 19  4 15.54 0.038040
#> 20  4 15.75 0.034660
#> 21  6 15.94 0.031580
#> 22  6 16.10 0.028780
#> 23  6 16.24 0.026220
#> 24  6 16.36 0.023890
#> 25  6 16.46 0.021770
#> 26  6 16.55 0.019840
#> 27  6 16.62 0.018070
#> 28  6 16.67 0.016470
#> 29  6 16.72 0.015010
#> 30  6 16.77 0.013670
#> 31  6 16.80 0.012460
#> 32  6 16.83 0.011350
#> 33  6 16.86 0.010340
#> 34  6 16.88 0.009424
#> 35  6 16.89 0.008587
#> 36  6 16.91 0.007824
#> 37  6 16.92 0.007129
#> 38  6 16.93 0.006495
#> 39  6 16.94 0.005918
#> 40  6 16.95 0.005393
#> 41  6 16.95 0.004914
#> 42  6 16.96 0.004477
#> 43  6 16.96 0.004079
#> 44  6 16.96 0.003717
#> 45  6 16.97 0.003387
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  28   9      0        180        1   107
#>   [2,]  33   2      0          2        0   150
#>   [3,]  35   5      1          2        0   172
#>   [4,]  35  10      1          9        0   106
#>   [5,]  34   5      0          5        0   120
#>   [6,]  35   2      0        180        0   121
#>   [7,]  37   9      0        180        1   151
#>   [8,]  38  13      1          0        1   161
#>   [9,]  38   2      0        115        0   150
#>  [10,]  36   1      0        180        1   155
#>  [11,]  35   0      0        180        1   119
#>  [12,]  38  12      1          8        1   120
#>  [13,]  36   5      1          0        1   115
#>  [14,]  33   6      1          1        1   115
#>  [15,]  38  16      1         10        0   160
#>  [16,]  38  12      1         11        1    92
#>  [17,]  40  12      1          9        0   153
#>  [18,]  37   1      1          0        1   146
#>  [19,]  40   2      1          1        1   148
#>  [20,]  42   2      0        180        1   100
#>  [21,]  38   5      1          3        0   125
#>  [22,]  40  11      1         10        1   120
#>  [23,]  42   2      0        180        0   100
#>  [24,]  43   4      1          0        1   130
#>  [25,]  42   4      0        180        0   162
#>  [26,]  42  15      1         13        1   125
#>  [27,]  40   3      1          1        0   170
#>  [28,]  42   2      0        180        1   124
#>  [29,]  44   5      1          1        0   170
#>  [30,]  41  13      1          1        0   140
#>  [31,]  41   5      1          4        1   141
#>  [32,]  44   2      1          1        1   150
#>  [33,]  43   2      0        180        1   140
#>  [34,]  45   2      0        180        1   140
#>  [35,]  47   4      1          3        0   118
#>  [36,]  48  15      0        180        1   160
#>  [37,]  45   3      0        150        0   130
#>  [38,]  44   3      1          0        1   180
#>  [39,]  46   7      1          2        0   166
#>  [40,]  43  29      0        180        1   180
#>  [41,]  45   4      1          0        0   124
#>  [42,]  43  10      0        180        0   185
#>  [43,]  46  13      1         10        0   100
#>  [44,]  44   0      1          0        1    96
#>  [45,]  47   4      1          3        1   160
#>  [46,]  43   3      1          0        1   124
#>  [47,]  45   8      1          0        1   117
#>  [48,]  46   2      1          1        1   122
#>  [49,]  44   4      1          0        1   114
#>  [50,]  47   2      0        180        0   108
#>  [51,]  46   5      1          3        0   130
#>  [52,]  44   2      0        180        0   142
#>  [53,]  45   9      1          0        1   145
#>  [54,]  47   3      1          1        1   120
#>  [55,]  48  12      1         11        0   200
#>  [56,]  47   5      1          3        1   130
#>  [57,]  47   9      1          6        0   170
#>  [58,]  46   3      1          0        1   119
#>  [59,]  49   4      0        180        0   117
#>  [60,]  47  10      0         10        1   140
#>  [61,]  48   2      1          0        0   184
#>  [62,]  50   4      1          1        0   125
#>  [63,]  50   6      1          2        1   140
#>  [64,]  46   3      1          1        1   140
#>  [65,]  50   7      0        180        1   110
#>  [66,]  49   2      0          2        0   105
#>  [67,]  49  15      1         11        1   160
#>  [68,]  49  23      0        179        1   112
#>  [69,]  52   2      0        180        1   170
#>  [70,]  50   7      1          0        1    92
#>  [71,]  51   3      1          2        0   113
#>  [72,]  50   1      1          0        0   150
#>  [73,]  50   9      0        180        0   130
#>  [74,]  49   7      1          4        1    90
#>  [75,]  47   8      0        180        0   160
#>  [76,]  47   6      0        180        1   162
#>  [77,]  51   8      0        180        1   140
#>  [78,]  46   1      1          1        0   145
#>  [79,]  48   7      1          0        1   110
#>  [80,]  53   8      0         36        1   160
#>  [81,]  48  17      1         10        0   111
#>  [82,]  47   2      1          1        0   110
#>  [83,]  49  15      0        180        1   160
#>  [84,]  54  17      1         12        1   102
#>  [85,]  53   5      0         77        0   159
#>  [86,]  54   6      1          3        0   129
#>  [87,]  51   3      1          1        0   140
#>  [88,]  50   2      0          5        1   106
#>  [89,]  50  14      1         13        0   170
#>  [90,]  48   3      1          2        0   150
#>  [91,]  49   5      1          2        1   150
#>  [92,]  52  14      1          7        1   200
#>  [93,]  48   6      0        180        0   160
#>  [94,]  48  11      1         10        0   120
#>  [95,]  51  13      0         99        1   160
#>  [96,]  54   9      1          0        1   138
#>  [97,]  49  16      0         16        0   125
#>  [98,]  55   3      1          1        0   150
#>  [99,]  54  23      1         10        0   131
#> [100,]  52   7      1          2        0   154
#> [101,]  55   6      1          2        1   114
#> [102,]  54   9      1          1        0   130
#> [103,]  52   4      0        180        1   180
#> [104,]  51  13      1         11        0   145
#> [105,]  54   4      1          0        1   121
#> [106,]  52   4      0        180        0   183
#> [107,]  55  28      1         13        1   160
#> [108,]  49   6      1          0        1   130
#> [109,]  49   1      0          1        1   110
#> [110,]  50   7      1          1        0   156
#> [111,]  53   8      1          0        1   130
#> [112,]  50   7      1          0        1   127
#> [113,]  52   5      0        175        1   117
#> [114,]  55   1      0        180        0   127
#> [115,]  55   2      0          2        0   145
#> [116,]  56   3      0        180        1   193
#> [117,]  56   2      0        180        0   132
#> [118,]  55   5      1          4        1   120
#> [119,]  52   8      0        180        0   119
#> [120,]  53  18      1          9        1   150
#> [121,]  54   3      0        180        1   180
#> [122,]  55   6      0        180        0   170
#> [123,]  52  16      0         16        0   152
#> [124,]  53  10      1          9        0   172
#> [125,]  52  16      1         14        0   170
#> [126,]  53   4      0        180        1   150
#> [127,]  55   6      0        180        1   100
#> [128,]  55   6      1          5        1   138
#> [129,]  54  12      1          0        1   190
#> [130,]  54   3      0        180        0   128
#> [131,]  56   3      0          8        1   139
#> [132,]  55   1      0          2        0   130
#> [133,]  57   3      0          3        0   120
#> [134,]  54   7      1          2        0   129
#> [135,]  54   2      1          1        0   135
#> [136,]  52   9      1          3        0   170
#> [137,]  57   5      1          3        1   138
#> [138,]  57   1      0          1        1   100
#> [139,]  56   4      1          0        1   140
#> [140,]  52   2      0        180        0   140
#> [141,]  55  11      1          7        0   104
#> [142,]  52  15      1         14        0   130
#> [143,]  56  14      1         11        0   130
#> [144,]  53   3      1          0        1   200
#> [145,]  54   5      0        180        1   108
#> [146,]  57   0      0          0        1   150
#> [147,]  53  21      1         13        1   130
#> [148,]  59   3      1          1        0   172
#> [149,]  58   6      1          0        1    90
#> [150,]  55   9      1          2        1   147
#> [151,]  55  13      0        166        1   140
#> [152,]  56   5      0          5        1   150
#> [153,]  54  23      1          8        0   120
#> [154,]  57   4      1          2        1   185
#> [155,]  53   7      1          0        1   120
#> [156,]  57  11      1         10        1   129
#> [157,]  55   5      0          5        1   131
#> [158,]  59  15      1         10        0   140
#> [159,]  58   1      1          1        1   200
#> [160,]  55   5      1          0        0   140
#> [161,]  56   7      1          5        1   120
#> [162,]  59   9      1          1        1   125
#> [163,]  57   1      0        180        0   148
#> [164,]  58   4      1          0        1   160
#> [165,]  60   5      1          1        0   138
#> [166,]  57   5      0        180        1   130
#> [167,]  58  11      1          9        1   124
#> [168,]  55   5      1          0        1   160
#> [169,]  57  10      1          9        0   103
#> [170,]  59   6      1          0        1   140
#> [171,]  59   5      0        180        1   155
#> [172,]  59   4      1          0        1   152
#> [173,]  58  26      1          0        1   189
#> [174,]  61   9      0          9        1   160
#> [175,]  58   4      1          3        0   120
#> [176,]  59   2      1          1        0   140
#> [177,]  58   8      0        161        1   140
#> [178,]  58  14      1          6        0   190
#> [179,]  61   4      1          3        0   151
#> [180,]  61   3      1          2        1   102
#> [181,]  58   1      0          1        1   100
#> [182,]  57   2      1          0        1   116
#> [183,]  58  10      0         10        1   150
#> [184,]  57  11      0        180        1   150
#> [185,]  61   3      0         17        0   143
#> [186,]  56  14      0         45        0   130
#> [187,]  57   3      1          2        0   120
#> [188,]  58  19      1         13        1   140
#> [189,]  56  13      1          6        1   158
#> [190,]  59   9      1          0        1    80
#> [191,]  55   4      1          3        1   160
#> [192,]  58  11      0        172        1   135
#> [193,]  60  12      1          0        1   114
#> [194,]  55   9      1          7        1   135
#> [195,]  61   4      1          0        1   115
#> [196,]  56   8      1          8        0   120
#> [197,]  61  13      1         12        1   130
#> [198,]  57   1      0          1        0   126
#> [199,]  57  15      1         13        1   110
#> [200,]  59   5      1          2        0   182
#> [201,]  58   5      1          1        1   135
#> [202,]  59  10      0        180        0   160
#> [203,]  62  10      1          0        1   153
#> [204,]  62   7      1          2        1   180
#> [205,]  57   3      1          0        0   100
#> [206,]  57   7      0        169        0   180
#> [207,]  61   7      0          7        1   150
#> [208,]  59  13      1          2        0   198
#> [209,]  57  12      1          9        1   120
#> [210,]  60  17      1          8        1   140
#> [211,]  58   3      1          0        1   146
#> [212,]  62   4      1          3        0   173
#> [213,]  58   2      0         30        0   202
#> [214,]  59   1      0        180        0   155
#> [215,]  63   6      0         28        1   120
#> [216,]  61  13      0         13        0   120
#> [217,]  57  18      1          9        1    93
#> [218,]  58  11      1          9        0   179
#> [219,]  62  17      1         10        1   180
#> [220,]  62   1      1          0        1   172
#> [221,]  63   3      1          1        0   180
#> [222,]  63   1      0        180        1   130
#> [223,]  61   7      0        180        0   135
#> [224,]  63  15      1         10        1   126
#> [225,]  64   4      0        180        0   130
#> [226,]  63   4      1          1        0   155
#> [227,]  59   8      0        180        1   140
#> [228,]  58   9      1          9        0   110
#> [229,]  62   7      0          7        0   150
#> [230,]  59   1      0         22        1   162
#> [231,]  59   4      0        180        0   196
#> [232,]  60   7      0          7        0   140
#> [233,]  59   5      1          1        0   148
#> [234,]  60   7      1          1        1    90
#> [235,]  63   1      0          1        0   130
#> [236,]  61  15      1         13        0   170
#> [237,]  59   4      0          4        0   149
#> [238,]  60   3      0          3        0   168
#> [239,]  64  10      1          9        0   160
#> [240,]  62   6      0          6        0   120
#> [241,]  63  12      1         10        0   200
#> [242,]  59  10      0        180        1   130
#> [243,]  60   8      0         17        1   130
#> [244,]  61   6      1          1        1   117
#> [245,]  64  12      1         11        0   160
#> [246,]  66   1      1          0        1   120
#> [247,]  64   6      1          0        1   140
#> [248,]  63  10      1          0        1   148
#> [249,]  63  14      1          9        0   123
#> [250,]  65  36      1         11        0   140
#> [251,]  66   3      1          1        0   127
#> [252,]  61  10      1          2        1   194
#> [253,]  64  32      1          9        1   160
#> [254,]  63  12      1          9        0   114
#> [255,]  63   7      0        180        0   120
#> [256,]  66   5      1          0        1   110
#> [257,]  65   8      1          0        0   168
#> [258,]  65  10      1          8        1   120
#> [259,]  64   0      0          0        1    90
#> [260,]  60   6      0        180        0   130
#> [261,]  61  12      1         11        0   154
#> [262,]  64   9      0        180        0   150
#> [263,]  61   4      0        180        1   113
#> [264,]  65   3      0        180        1   190
#> [265,]  66   6      1          1        1   130
#> [266,]  63  12      0         12        1   150
#> [267,]  63   5      1          4        0   170
#> [268,]  63   2      1          1        0   180
#> [269,]  62  13      1         11        0   180
#> [270,]  64   2      0          2        0   201
#> [271,]  66  18      1          5        0   142
#> [272,]  61  14      1          5        0   140
#> [273,]  63   2      1          0        0   140
#> [274,]  65   8      1          0        1   140
#> [275,]  64  13      1         12        1   150
#> [276,]  64   6      1          0        1   125
#> [277,]  66  13      1          0        0   118
#> [278,]  64  14      1         13        1   150
#> [279,]  65   3      0          3        0   105
#> [280,]  67   4      1          3        0   130
#> [281,]  66   3      1          0        1   135
#> [282,]  65   2      1          1        1   170
#> [283,]  68   1      0        180        1   166
#> [284,]  67   8      1          1        1   130
#> [285,]  68   5      0          5        1    90
#> [286,]  66  14      0        180        0   130
#> [287,]  68  18      0        180        1   260
#> [288,]  63  10      0         18        1   130
#> [289,]  68  11      0        180        0   160
#> [290,]  68  14      0         79        0   172
#> [291,]  66  12      1         10        1   150
#> [292,]  65  15      1         12        1   150
#> [293,]  66  11      1          0        0   100
#> [294,]  63   2      0        180        0   150
#> [295,]  65  11      1          6        0   130
#> [296,]  69  21      1         10        0   180
#> [297,]  69   6      0        180        1   100
#> [298,]  66   9      1          8        0   130
#> [299,]  68  14      1         13        1   140
#> [300,]  69   1      1          0        0   170
#> [301,]  67   1      0        180        1   160
#> [302,]  67   7      1          4        1   130
#> [303,]  63   2      1          0        0    99
#> [304,]  67   2      0        180        0   184
#> [305,]  65  10      1          1        1   148
#> [306,]  66  19      1         12        1   150
#> [307,]  69   6      0         99        1   140
#> [308,]  65   4      1          1        0   130
#> [309,]  64   4      0        179        0   160
#> [310,]  66   4      0        180        1   130
#> [311,]  70  15      1         12        1   132
#> [312,]  64   0      1          0        1   118
#> [313,]  66   7      1          5        1   131
#> [314,]  66   4      0        180        0   177
#> [315,]  68   4      1          0        1   160
#> [316,]  69   4      1          3        1   150
#> [317,]  69  17      1         10        0   140
#> [318,]  69   8      0         93        0   140
#> [319,]  64  21      0         21        1   155
#> [320,]  68  18      1          0        1   160
#> [321,]  65   6      0        101        1   115
#> [322,]  71   3      0          5        0   112
#> [323,]  68   7      0        150        0   210
#> [324,]  67   2      0        180        0   128
#> [325,]  66   9      1          3        1   151
#> [326,]  70   4      1          0        1   180
#> [327,]  69   8      0        180        1   153
#> [328,]  70  14      0        171        0   166
#> [329,]  67   6      1          4        0   130
#> [330,]  68   7      1          0        1   150
#> [331,]  67  14      1         13        0   130
#> [332,]  65  14      1         13        1   150
#> [333,]  71   3      0        103        0   133
#> [334,]  69   3      0          3        1   130
#> [335,]  70  22      1         13        0   103
#> [336,]  67   1      0         36        1   104
#> [337,]  68   6      0        180        0   145
#> [338,]  69   8      1          5        1   195
#> [339,]  69   6      1          4        1   174
#> [340,]  69   8      1          7        1   108
#> [341,]  67   3      0        180        0   110
#> [342,]  69  19      0        180        0   130
#> [343,]  67  14      0        172        1   140
#> [344,]  66   2      0        180        0   130
#> [345,]  67   7      1          4        0   122
#> [346,]  69   4      1          3        0   132
#> [347,]  68   2      0          7        1   130
#> [348,]  69   8      1          2        0   121
#> [349,]  67  13      1          9        0   130
#> [350,]  70   3      0        123        0   130
#> [351,]  72   5      1          4        0   170
#> [352,]  67  22      1          1        1   140
#> [353,]  68   3      0         19        0   135
#> [354,]  67  12      1          8        0   120
#> [355,]  69   1      0          1        1   110
#> [356,]  67   1      0          1        1    60
#> [357,]  67   4      0         60        1   136
#> [358,]  67   8      1          0        1   130
#> [359,]  68  10      1          8        1   160
#> [360,]  66  24      1         13        0   130
#> [361,]  72  30      1          0        1   145
#> [362,]  71   6      0          9        0   120
#> [363,]  69  10      1          6        1   120
#> [364,]  70  11      0        180        1   210
#> [365,]  72  19      1          8        0   120
#> [366,]  72  12      1         10        0   170
#> [367,]  67   8      0        180        1   170
#> [368,]  67   9      0        180        0   158
#> [369,]  73  13      0        152        1   130
#> [370,]  70   5      0        180        0   150
#> [371,]  72   2      0          2        1   100
#> [372,]  67   4      1          1        0   134
#> [373,]  72   6      1          5        0   115
#> [374,]  70   3      0        180        0   121
#> [375,]  69   3      0        180        0   220
#> [376,]  73   6      0        180        1   117
#> [377,]  69  16      1         10        1   140
#> [378,]  69   8      1          1        0   164
#> [379,]  68   7      0        180        1   130
#> [380,]  72  16      1          1        1   130
#> [381,]  70   4      0        180        0   180
#> [382,]  73   6      1          0        1   270
#> [383,]  72   8      1          1        1   150
#> [384,]  71   2      1          0        1   180
#> [385,]  68  15      1         13        1   130
#> [386,]  70   3      0          3        1   159
#> [387,]  70  13      1          9        0   100
#> [388,]  72   6      0        180        1   130
#> [389,]  74   8      1          0        1    85
#> [390,]  73   4      0        180        1   154
#> [391,]  71   3      1          1        0   150
#> [392,]  74  20      0         20        1   180
#> [393,]  68   9      0        180        1   120
#> [394,]  71  20      1         10        0   140
#> [395,]  74   0      1          0        1    90
#> [396,]  70   5      1          0        1   190
#> [397,]  71  17      1         11        0   160
#> [398,]  71   8      1          7        0   149
#> [399,]  71   3      1          2        1   190
#> [400,]  73  10      1          8        0   106
#> [401,]  70  26      1         11        1   120
#> [402,]  74   4      0          4        0   120
#> [403,]  72   5      1          3        1   160
#> [404,]  73   6      0        180        0   110
#> [405,]  72  15      1          0        1   150
#> [406,]  72   8      1          0        1   140
#> [407,]  73  17      1         11        0   140
#> [408,]  71  13      1          8        0   121
#> [409,]  70   4      1          0        1   140
#> [410,]  71  14      1         13        1   170
#> [411,]  74   7      1          0        1   117
#> [412,]  72  10      1          8        1   153
#> [413,]  69   7      0        180        1   144
#> [414,]  70   8      0          8        0   120
#> [415,]  71  10      1          9        1   120
#> [416,]  75   2      1          1        0   145
#> [417,]  73  10      1          9        1   146
#> [418,]  72  10      1          9        1   160
#> [419,]  73  10      1         10        1   120
#> [420,]  75   9      1          7        0   140
#> [421,]  75  13      1          1        1   130
#> [422,]  71   4      0          4        0   134
#> [423,]  70   7      1          4        0   184
#> [424,]  72   1      1          1        0   168
#> [425,]  70   3      0          3        0   150
#> [426,]  73   5      1          3        1   112
#> [427,]  72   2      0        180        0   120
#> [428,]  72   4      1          0        1   197
#> [429,]  71   3      1          0        0   144
#> [430,]  73   5      0        180        0   126
#> [431,]  73   4      0        180        0   124
#> [432,]  74  34      1          8        1   233
#> [433,]  76   3      1          0        1   120
#> [434,]  72   3      0        180        0   160
#> [435,]  76   5      0          5        1   130
#> [436,]  77  11      0         11        1   150
#> [437,]  77   4      0          4        0   185
#> [438,]  75   3      1          1        0   180
#> [439,]  73  15      0         15        1   160
#> [440,]  71  16      0        180        0   140
#> [441,]  74   7      0        180        1   150
#> [442,]  76   1      0        180        0   114
#> [443,]  74   2      1          1        0   140
#> [444,]  76   8      1          0        1   141
#> [445,]  73   6      0          6        1   114
#> [446,]  75  23      1         14        1   110
#> [447,]  72   4      0         85        1   120
#> [448,]  72   4      1          3        0   160
#> [449,]  76  17      1          0        1   200
#> [450,]  73   4      1          3        1   125
#> [451,]  76  13      1         10        0   110
#> [452,]  75   4      1          0        1   122
#> [453,]  75   7      0          7        0   190
#> [454,]  75   0      0          0        1   130
#> [455,]  74   8      1          0        1   105
#> [456,]  76  13      1          8        1   148
#> [457,]  75   4      1          2        1   188
#> [458,]  76   4      0          4        1   155
#> [459,]  75   1      0          1        1   125
#> [460,]  73   1      0         52        1   105
#> [461,]  78  12      1         11        1   160
#> [462,]  76  44      1         10        0   105
#> [463,]  74  10      1          0        1   135
#> [464,]  76   5      1          0        1   167
#> [465,]  75   9      0        180        1   140
#> [466,]  77  12      1          9        1   100
#> [467,]  73  10      1          9        0   146
#> [468,]  77  12      0        180        0   130
#> [469,]  76  12      1         11        1   120
#> [470,]  78   5      1          0        1   170
#> [471,]  75   6      0        180        0   150
#> [472,]  78  18      0         18        1   144
#> [473,]  77   3      0        180        0   110
#> [474,]  76  29      0         47        0    90
#> [475,]  73  11      1          2        1   110
#> [476,]  74   2      0        180        0   100
#> [477,]  78  32      1          9        1   198
#> [478,]  79   6      0        180        0   170
#> [479,]  78  13      1          5        0   130
#> [480,]  75  12      1          1        1   120
#> [481,]  78  15      0        180        1   270
#> [482,]  80   8      0          8        1   120
#> [483,]  75  13      1          6        0   150
#> [484,]  74  10      1          8        0   135
#> [485,]  76   1      0          1        1    83
#> [486,]  78  12      1          9        0   150
#> [487,]  78   2      1          1        0   130
#> [488,]  77   2      1          0        1   143
#> [489,]  78  10      0        180        1   130
#> [490,]  76  11      1          0        0   120
#> [491,]  76   7      0         29        1   150
#> [492,]  77  24      0         24        1   160
#> [493,]  79   8      0         32        1   120
#> [494,]  80   6      0          6        1   150
#> [495,]  78   6      1          0        1   240
#> [496,]  76   3      1          0        1   140
#> [497,]  78  11      1          1        1   140
#> [498,]  79   2      1          0        1   121
#> [499,]  78  14      1          0        1   140
#> [500,]  81   1      0          1        0   130
#> [501,]  78  11      1          8        1   118
#> [502,]  76   4      0          4        1   160
#> [503,]  79   4      0          4        1   125
#> [504,]  76  10      1          8        0   180
#> [505,]  76  12      1         10        1   127
#> [506,]  80   3      1          0        1   120
#> [507,]  75   2      1          1        1   204
#> [508,]  76   1      0          1        1   140
#> [509,]  76   1      0          1        1    90
#> [510,]  78   7      1          0        1   110
#> [511,]  79   3      0          3        0   120
#> [512,]  79   4      1          0        1   120
#> [513,]  81   1      0        180        0   120
#> [514,]  77   9      1          4        0   141
#> [515,]  82   5      0          8        1   120
#> [516,]  78   4      0         59        1   112
#> [517,]  76   7      0        161        0   151
#> [518,]  79  10      0         10        1   120
#> [519,]  80  15      1          0        1    90
#> [520,]  80   6      0        173        1   160
#> [521,]  79   1      0         37        1   140
#> [522,]  78   7      0          7        1   147
#> [523,]  77  13      1          0        1   190
#> [524,]  80   5      1          1        1   108
#> [525,]  78   4      0        180        0   175
#> [526,]  79   3      0          3        1   101
#> [527,]  78  26      1          5        0   194
#> [528,]  76   1      0        166        0   131
#> [529,]  81   4      1          1        1   104
#> [530,]  78  20      1          0        1   109
#> [531,]  80   1      0          1        0   100
#> [532,]  78   3      1          1        1   152
#> [533,]  77  10      1          8        1   130
#> [534,]  82   3      1          1        1   144
#> [535,]  77   5      0         85        0   188
#> [536,]  80   6      1          0        1   119
#> [537,]  82   1      1          0        1    82
#> [538,]  79  10      0        180        1   150
#> [539,]  77   4      0        180        1    98
#> [540,]  81   1      0        108        0   129
#> [541,]  78  12      0        180        0   134
#> [542,]  79   1      0        125        0   193
#> [543,]  82  21      1          2        0   155
#> [544,]  84  22      1         10        0   180
#> [545,]  79   4      0          4        1   121
#> [546,]  80   6      0          6        1   110
#> [547,]  83   9      1          5        1   170
#> [548,]  83   4      0        103        0    97
#> [549,]  81  11      1          8        0   160
#> [550,]  80  11      1          8        0   170
#> [551,]  78  23      1         10        1   145
#> [552,]  79   4      0          4        1   183
#> [553,]  78   9      1          4        1   120
#> [554,]  81  15      0        180        1   140
#> [555,]  80   7      1          0        1   146
#> [556,]  84   5      1          1        1    85
#> [557,]  81  20      1          9        0   170
#> [558,]  83   8      0          8        0   115
#> [559,]  81  16      0         16        1   110
#> [560,]  80   6      1          0        1   150
#> [561,]  81   8      0        180        0   146
#> [562,]  80   8      1          7        0   160
#> [563,]  79   0      1          0        1    96
#> [564,]  81   2      1          1        0   198
#> [565,]  83   2      0          2        1   155
#> [566,]  80   3      1          1        1   230
#> [567,]  82  23      1          0        0   110
#> [568,]  84   4      0          4        1    85
#> [569,]  81   1      0          1        1   150
#> [570,]  85   3      1          2        1   160
#> [571,]  80  13      1          8        1   140
#> [572,]  84   4      0         89        1   129
#> [573,]  80   2      1          0        1   130
#> [574,]  80   6      0         71        1   189
#> [575,]  82  19      0         19        0   120
#> [576,]  80  30      1         13        0   220
#> [577,]  83   9      0        180        0   198
#> [578,]  79  14      1          0        0   110
#> [579,]  83   3      0        114        0    98
#> [580,]  81  14      1         12        1   128
#> [581,]  83   2      0        154        0   130
#> [582,]  83   1      0        180        0   160
#> [583,]  81   1      0          1        1   145
#> [584,]  81  12      0         12        1   163
#> [585,]  82   5      1          0        1   146
#> [586,]  81   4      0          4        0   160
#> [587,]  86  12      0        180        1   120
#> [588,]  81  19      1         14        0   120
#> [589,]  82   3      1          2        0   130
#> [590,]  82  15      1          0        0   183
#> [591,]  83   7      0        126        0   135
#> [592,]  81  16      1          9        0   180
#> [593,]  84   6      0        165        0   145
#> [594,]  82   9      0        180        1   134
#> [595,]  81  13      0        180        0   152
#> [596,]  85   3      0          3        1   118
#> [597,]  81   2      1          0        1   118
#> [598,]  83   9      0        180        1   149
#> [599,]  83   4      0          4        0   130
#> [600,]  87   2      0          5        1   137
#> [601,]  86  12      1          0        1   132
#> [602,]  86   6      1          0        1   140
#> [603,]  84  16      0         70        1   150
#> [604,]  83  10      1          0        1   190
#> [605,]  86   2      0        180        1   169
#> [606,]  88  14      1          3        1   130
#> [607,]  84   3      0          3        1   121
#> [608,]  87   2      0        180        0   113
#> [609,]  84   3      0        180        1   170
#> [610,]  86   4      0         38        1   122
#> [611,]  82   4      0          4        0   130
#> [612,]  86  13      0        177        0   163
#> [613,]  85   3      0          3        1   113
#> [614,]  86   6      0          6        1   117
#> [615,]  88   4      0          4        0   100
#> [616,]  83  20      1          3        1   150
#> [617,]  88   4      0          4        1   115
#> [618,]  85  22      0         22        1   184
#> [619,]  86   9      1          7        1   142
#> [620,]  88   2      0        180        1    68
#> [621,]  87   8      0          8        1   157
#> [622,]  88   4      0          4        0    86
#> [623,]  89   4      0          4        1   153
#> [624,]  89   5      0        119        1   140
#> [625,]  87   1      0          1        0   170
#> [626,]  85   8      0          8        1   136
#> [627,]  84   0      0        180        1   136
#> [628,]  89  10      0         46        1   170
#> [629,]  88   1      0          1        0   135
#> [630,]  86   4      0        180        1   145
#> [631,]  87   2      0        180        0   160
#> [632,]  87   6      1          0        0   125
#> [633,]  91  10      0        145        0   135
#> [634,]  88   7      0         24        0   119
#> [635,]  88   8      0         50        1   154
#> [636,]  87   6      0        126        1   168
#> [637,]  86   9      1          7        0   130
#> [638,]  90   4      1          0        0   121
#> [639,]  87   5      0         36        1   150
#> [640,]  90   5      1          0        1   125
#> [641,]  89   3      1          1        1   160
#> [642,]  92   1      0          1        1   167
#> [643,]  91   3      0         33        1   137
#> [644,]  88   5      0        158        0   100
#> [645,]  87   7      0         74        1   105
#> [646,]  89  12      1          0        1   130
#> [647,]  89   2      0        168        0   118
#> [648,]  91   5      0        169        1   176
#> [649,]  89  52      0         52        1   130
#> [650,]  92   7      0          7        1   110
#> [651,]  91   0      0          0        0     0
#> [652,]  91   4      1          0        1   120
#> [653,]  90   1      0          1        1   118
#> [654,]  91   2      0          2        1   116
#> [655,]  93   8      0        179        1   110
#> [656,]  94   8      0          8        1   142
#> [657,]  92   4      0         76        1   149
#> [658,]  91   1      0        180        0   158
#> [659,]  90  16      0         16        1   106
#> [660,]  90   3      0         67        0   162
#> [661,]  96   3      0         12        1    97
#> [662,]  95   8      1          5        1   150
#> [663,]  94   3      0         26        1   144
#> [664,]  91  12      0         53        1   212
#> [665,]  91   7      0          7        0   135
#> [666,]  93   0      1          0        1   122
#> [667,]  92   5      0         69        0   139
#> [668,]  92   2      0          2        0   112
#> [669,]  96   3      1          0        1   104
#> [670,]  96  15      1          0        1   140
#> 
#> $y
#>   [1] 180.0+   2.0+   5.0+ 180.0+   5.0+ 180.0+ 180.0+ 180.0+ 115.0  180.0+
#>  [11] 180.0+  12.0    5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 180.0+
#>  [21]   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 155.0+ 180.0+
#>  [31]   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 150.0  180.0+ 180.0+ 180.0+
#>  [41] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 161.0+ 180.0+ 180.0+
#>  [51]   5.0+ 180.0+ 177.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  10.0+
#>  [61] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0  179.0+ 179.0+ 180.0+ 180.0+
#>  [71] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   7.0   36.0 
#>  [81]  88.0+ 180.0+ 180.0+ 180.0+  77.0  180.0+ 180.0+   5.0  180.0+ 180.0+
#>  [91] 180.0+  85.0  180.0+ 180.0+  99.0  180.0+  16.0+ 180.0+ 152.0+   7.0+
#> [101]   6.0+ 180.0+ 180.0+  13.0+ 180.0+ 180.0+  28.0    6.0+   1.0  180.0+
#> [111] 180.0+ 180.0+ 175.0+ 180.0+   2.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [121] 180.0+ 180.0+  16.0+ 180.0+  16.0  180.0+ 180.0+ 180.0+  12.0+ 180.0+
#> [131]   8.0    2.0    3.0+ 180.0+ 180.0+ 180.0+ 140.0    1.0  165.0  180.0+
#> [141] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   0.5  180.0+ 180.0+ 180.0+  15.0 
#> [151] 166.0+   5.0+ 180.0+   4.0+ 180.0+ 180.0+   5.0+ 180.0+   1.0  180.0+
#> [161] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  64.0 
#> [171] 180.0+ 180.0+ 180.0+   9.0+ 180.0+ 180.0+ 161.0+ 171.0+ 180.0+   3.0 
#> [181]   1.0  180.0+  10.0+ 180.0+  17.0   45.0    3.0+  19.0  180.0+   9.0+
#> [191] 180.0+ 172.0+ 172.0+  24.0  180.0+   8.0  180.0+   1.0+  15.0  180.0+
#> [201] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 169.0    7.0  180.0+ 180.0+ 180.0+
#> [211]   3.0+ 180.0+  30.0  180.0+  28.0   13.0+  18.0  180.0+ 180.0+   1.0 
#> [221] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+   9.0    7.0+  22.0 
#> [231] 180.0+   7.0+ 180.0+ 180.0+   1.0  180.0+   4.0+   3.0+ 167.0    6.0+
#> [241] 180.0+ 180.0+  17.0  180.0+  12.0  180.0+ 180.0+ 180.0+  14.0+  36.0 
#> [251]   3.0+  88.0  180.0+  12.0  180.0+ 180.0+ 180.0+ 180.0+   0.5  180.0+
#> [261]  12.0+ 180.0+ 180.0+ 180.0+ 180.0+  12.0  180.0+ 180.0+ 180.0+   2.0+
#> [271]  18.0+ 180.0+   2.0+  15.0   13.0  180.0+ 166.0+  14.0+   3.0  180.0+
#> [281]   3.0+ 175.0+ 180.0+   8.0    5.0  180.0+ 180.0+  18.0  180.0+  79.0 
#> [291]  80.0   15.0+ 180.0+ 180.0+ 180.0+ 174.0+ 180.0+ 180.0+ 180.0+ 175.0 
#> [301] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  19.0+  99.0  180.0+ 179.0+ 180.0+
#> [311] 180.0+   0.5    7.0+ 180.0+ 180.0+ 152.0+ 180.0+  93.0   21.0+  18.0+
#> [321] 101.0    5.0  150.0  180.0+ 180.0+ 180.0+ 180.0+ 171.0    6.0  180.0+
#> [331] 180.0+  14.0+ 103.0    3.0+ 180.0+  36.0  180.0+ 180.0+  97.0    8.0+
#> [341] 180.0+ 180.0+ 172.0+ 180.0+   7.0  180.0+   7.0    8.0+  13.0+ 123.0 
#> [351] 180.0+  51.0   19.0  180.0+   1.0    1.0   60.0  180.0+  10.0+ 180.0+
#> [361] 162.0    9.0  180.0+ 180.0+ 180.0+  12.0  180.0+ 180.0+ 152.0  180.0+
#> [371]   2.0   76.0  180.0+ 180.0+ 180.0+ 180.0+  16.0+ 180.0+ 180.0+  16.0+
#> [381] 180.0+   6.0  180.0+ 180.0+  15.0    3.0+  13.0+ 180.0+ 180.0+ 180.0+
#> [391]   3.0+  20.0  180.0+  20.0    0.5  180.0+ 180.0+   8.0    3.0   87.0 
#> [401] 180.0+   4.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 175.0  180.0+  14.0+
#> [411] 180.0+  10.0+ 180.0+   8.0+ 179.0+ 180.0+ 180.0+ 159.0   15.0  180.0+
#> [421]  13.0    4.0+ 104.0+   1.0    3.0+   5.0  180.0+ 180.0+ 180.0+ 180.0+
#> [431] 180.0+  34.0  180.0+ 180.0+   5.0   11.0+   4.0+ 180.0+  15.0+ 180.0+
#> [441] 180.0+ 180.0+ 180.0+ 180.0+   6.0  180.0+  85.0  180.0+  17.0+ 180.0+
#> [451] 174.0+   4.0    7.0    0.5  180.0+ 180.0+  46.0    4.0    1.0   52.0 
#> [461]  12.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  12.0  180.0+
#> [471] 180.0+  18.0  180.0+  47.0   11.0  180.0+  32.0  180.0+ 172.0   12.0 
#> [481] 180.0+   8.0  180.0+ 180.0+   1.0  180.0+ 180.0+   2.0  180.0+  11.0 
#> [491]  29.0   24.0   32.0    6.0  180.0+   3.0+ 180.0+ 180.0+ 180.0+   1.0 
#> [501]  11.0    4.0    4.0   10.0+ 180.0+   3.0+   2.0+   1.0    1.0   43.0 
#> [511]   3.0  138.0  180.0+  71.0    8.0   59.0  161.0   10.0+ 180.0+ 173.0 
#> [521]  37.0    7.0+  22.0    5.0+ 180.0+   3.0  171.0+ 166.0+  71.0   20.0+
#> [531]   1.0    3.0+  10.0  180.0+  85.0    6.0    1.0  180.0+ 180.0+ 108.0 
#> [541] 180.0+ 125.0  180.0+ 180.0+   4.0    6.0    9.0+ 103.0  180.0+ 169.0 
#> [551]  70.0    4.0  180.0+ 180.0+   7.0+ 180.0+  20.0    8.0+  16.0  180.0+
#> [561] 180.0+ 180.0+   0.5  180.0+   2.0    3.0+  62.0    4.0    1.0  180.0+
#> [571] 180.0+  89.0  180.0+  71.0   19.0   30.0  180.0+ 180.0+ 114.0  180.0+
#> [581] 154.0  180.0+   1.0   12.0    5.0+   4.0+ 180.0+ 180.0+   3.0   83.0 
#> [591] 126.0  180.0+ 165.0  180.0+ 180.0+   3.0+ 180.0+ 180.0+   4.0+   5.0 
#> [601] 180.0+   6.0   70.0  180.0+ 180.0+  14.0    3.0  180.0+ 180.0+  38.0 
#> [611]   4.0  177.0    3.0+   6.0+   4.0+  20.0    4.0   22.0   11.0  180.0+
#> [621]   8.0+   4.0    4.0  119.0    1.0+   8.0  180.0+  46.0    1.0+ 180.0+
#> [631] 180.0+  25.0  145.0   24.0   50.0  126.0  180.0+   4.0   36.0   89.0 
#> [641]   3.0+   1.0   33.0  158.0   74.0  180.0+ 168.0  169.0   52.0    7.0 
#> [651]   0.5    4.0    1.0+   2.0  179.0+   8.0+  76.0  180.0+  16.0   67.0 
#> [661]  12.0    8.0   26.0   53.0    7.0+   0.5   69.0    2.0    3.0   15.0+
#> 
#> $weights
#> NULL
#> 
#> $offset
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Warning: Multiple lambdas have been fit. Lambda will be set to 0.01 (see parameter 's').

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8659554