GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.coxnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.coxnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Caution: This learner is different to learners calling glmnet::cv.glmnet() in that it does not use the internal optimization of parameter lambda. Instead, lambda needs to be tuned by the user (e.g., via mlr3tuning). When lambda is tuned, the glmnet will be trained for each tuning iteration. While fitting the whole path of lambdas would be more efficient, as is done by default in glmnet::glmnet(), tuning/selecting the parameter at prediction time (using parameter s) is currently not supported in mlr3 (at least not in efficient manner). Tuning the s parameter is, therefore, currently discouraged.

When the data are i.i.d. and efficiency is key, we recommend using the respective auto-tuning counterpart in mlr_learners_surv.cv_glmnet(). However, in some situations this is not applicable, usually when data are imbalanced or not i.i.d. (longitudinal, time-series) and tuning requires custom resampling strategies (blocked design, stratification).

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.coxnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exact	logical	FALSE	TRUE, FALSE	-

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	TRUE	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	0.01		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.logistic	character	Newton	Newton, modified.Newton	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvGlmnet

Methods

Public methods

LearnerSurvGlmnet$new()

LearnerSurvGlmnet$selected_features()

LearnerSurvGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.glmnet")
print(learner)
#> 
#> ── <LearnerSurvGlmnet> (surv.glmnet): Regularized Generalized Linear Model ─────
#> • Model: -
#> • Parameters: use_pred_offset=TRUE
#> • Packages: mlr3, mlr3proba, mlr3extralearners, and glmnet
#> • Predict Types: [crank], distr, and lp
#> • Feature Types: logical, integer, and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: offset, selected_features, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#>    Df  %Dev   Lambda
#> 1   0  0.00 0.189500
#> 2   1  0.60 0.172600
#> 3   2  1.13 0.157300
#> 4   2  1.98 0.143300
#> 5   2  2.72 0.130600
#> 6   3  5.38 0.119000
#> 7   3  7.72 0.108400
#> 8   3  9.48 0.098790
#> 9   3 10.85 0.090020
#> 10  3 11.95 0.082020
#> 11  3 12.85 0.074730
#> 12  3 13.59 0.068090
#> 13  3 14.21 0.062040
#> 14  3 14.73 0.056530
#> 15  4 15.18 0.051510
#> 16  4 15.56 0.046930
#> 17  4 15.89 0.042770
#> 18  4 16.16 0.038970
#> 19  6 16.42 0.035500
#> 20  6 16.65 0.032350
#> 21  6 16.85 0.029480
#> 22  6 17.01 0.026860
#> 23  6 17.15 0.024470
#> 24  6 17.27 0.022300
#> 25  6 17.37 0.020320
#> 26  6 17.45 0.018510
#> 27  6 17.52 0.016870
#> 28  6 17.58 0.015370
#> 29  6 17.63 0.014000
#> 30  6 17.67 0.012760
#> 31  6 17.71 0.011630
#> 32  6 17.74 0.010590
#> 33  6 17.76 0.009652
#> 34  6 17.78 0.008795
#> 35  6 17.80 0.008013
#> 36  6 17.81 0.007302
#> 37  6 17.83 0.006653
#> 38  6 17.84 0.006062
#> 39  6 17.85 0.005523
#> 40  6 17.85 0.005033
#> 41  6 17.86 0.004586
#> 42  6 17.86 0.004178
#> 43  6 17.87 0.003807
#> 44  6 17.87 0.003469
#> 45  6 17.87 0.003161
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  32   5      1          0        1   121
#>   [2,]  35   5      1          2        0   172
#>   [3,]  35   2      0        180        0   121
#>   [4,]  35   2      1          1        1   112
#>   [5,]  38   2      0        115        0   150
#>   [6,]  36   1      0        180        1   155
#>   [7,]  35   0      0        180        1   119
#>   [8,]  36   5      1          0        1   115
#>   [9,]  33   6      1          1        1   115
#>  [10,]  38  16      1         10        0   160
#>  [11,]  40  12      1          9        0   153
#>  [12,]  42   3      1          1        1   130
#>  [13,]  40   2      1          1        1   148
#>  [14,]  42   2      0        180        1   100
#>  [15,]  38   5      1          3        0   125
#>  [16,]  40  11      1         10        1   120
#>  [17,]  42   2      0        180        0   100
#>  [18,]  43   3      1          0        1   100
#>  [19,]  41   2      1          1        0   166
#>  [20,]  40   1      1          0        1   145
#>  [21,]  43   4      1          0        1   130
#>  [22,]  42   4      0        180        0   162
#>  [23,]  42  15      1         13        1   125
#>  [24,]  44   5      1          1        0   170
#>  [25,]  45   3      0        180        1   154
#>  [26,]  41  10      1          8        0   150
#>  [27,]  44   3      0        180        0   141
#>  [28,]  41  13      1          1        0   140
#>  [29,]  45   9      1          7        0   110
#>  [30,]  45   6      0        180        1   170
#>  [31,]  44   2      1          1        1   150
#>  [32,]  43   2      0        180        1   140
#>  [33,]  47   4      1          3        0   118
#>  [34,]  45   3      0        150        0   130
#>  [35,]  44   3      1          0        1   180
#>  [36,]  46   7      1          2        0   166
#>  [37,]  43  29      0        180        1   180
#>  [38,]  45   4      1          0        0   124
#>  [39,]  43  10      0        180        0   185
#>  [40,]  47   6      1          0        1   116
#>  [41,]  44   0      1          0        1    96
#>  [42,]  45   8      1          0        1   117
#>  [43,]  45   5      0          5        0   141
#>  [44,]  46   2      1          1        1   122
#>  [45,]  46   6      1          0        1   100
#>  [46,]  44   4      1          0        1   114
#>  [47,]  47   2      0        180        0   108
#>  [48,]  44   9      1          8        1   135
#>  [49,]  45   5      0        180        1   190
#>  [50,]  46   4      0        180        1   121
#>  [51,]  45   9      1          0        1   145
#>  [52,]  47   3      1          1        1   120
#>  [53,]  48   3      0        180        0   154
#>  [54,]  47   5      1          3        1   130
#>  [55,]  46   3      1          0        1   119
#>  [56,]  50   1      1          0        1   129
#>  [57,]  48   2      1          0        0   184
#>  [58,]  47   7      0        180        0   145
#>  [59,]  50   4      1          1        0   125
#>  [60,]  50   6      1          2        1   140
#>  [61,]  49   7      1          7        1   110
#>  [62,]  46   3      1          1        1   140
#>  [63,]  50   7      0        180        1   110
#>  [64,]  51   1      0          1        1   145
#>  [65,]  47   2      0        180        0   150
#>  [66,]  46   6      1          0        1   156
#>  [67,]  52   2      0        180        1   170
#>  [68,]  50   7      1          0        1    92
#>  [69,]  50   4      0          4        1   100
#>  [70,]  51   3      1          2        0   113
#>  [71,]  50   1      1          0        0   150
#>  [72,]  50   9      0        180        0   130
#>  [73,]  49   7      1          4        1    90
#>  [74,]  51   8      0        180        1   140
#>  [75,]  52   2      0        180        0   155
#>  [76,]  46   3      0        180        1   120
#>  [77,]  46   1      1          1        0   145
#>  [78,]  50   4      1          1        0   150
#>  [79,]  47   2      1          1        0   110
#>  [80,]  52   4      1          4        0   152
#>  [81,]  49   9      1          3        0   102
#>  [82,]  49  15      0        180        1   160
#>  [83,]  53   5      0        180        1   140
#>  [84,]  54  17      1         12        1   102
#>  [85,]  53   7      1          0        0   199
#>  [86,]  54   6      1          3        0   129
#>  [87,]  51   3      1          1        0   140
#>  [88,]  50   2      0          5        1   106
#>  [89,]  50  14      1         13        0   170
#>  [90,]  53   8      1          7        0   160
#>  [91,]  51  25      1          1        0   202
#>  [92,]  53   4      0          4        0   140
#>  [93,]  52  14      1          7        1   200
#>  [94,]  48  11      1         10        0   120
#>  [95,]  53   4      1          0        1   156
#>  [96,]  49  16      0         16        0   125
#>  [97,]  54  23      1         10        0   131
#>  [98,]  52   7      1          2        0   154
#>  [99,]  55   6      1          2        1   114
#> [100,]  54   9      1          1        0   130
#> [101,]  55   4      1          2        0   150
#> [102,]  52   4      0        180        1   180
#> [103,]  50   5      1          4        1   150
#> [104,]  54   4      1          0        1   121
#> [105,]  50   3      0        174        1   153
#> [106,]  55  28      1         13        1   160
#> [107,]  49   6      1          0        1   130
#> [108,]  50   7      1          1        0   156
#> [109,]  53   9      0          9        1    95
#> [110,]  50   7      1          0        1   127
#> [111,]  55   2      0          2        0   145
#> [112,]  54   7      1          0        1   100
#> [113,]  56   2      0        180        0   132
#> [114,]  55   5      1          4        1   120
#> [115,]  55   6      0        180        0   170
#> [116,]  52  16      1         14        0   170
#> [117,]  55   6      0        180        1   100
#> [118,]  55   6      1          5        1   138
#> [119,]  54  12      1          0        1   190
#> [120,]  55   2      0        134        1   140
#> [121,]  54   3      0        180        0   128
#> [122,]  56   3      0          8        1   139
#> [123,]  57   3      0          3        0   120
#> [124,]  54   2      1          1        1   176
#> [125,]  57   1      0        180        1   156
#> [126,]  57   1      0          1        1   100
#> [127,]  52   2      0        180        0   140
#> [128,]  55  11      1          7        0   104
#> [129,]  56  14      1         11        0   130
#> [130,]  57  10      0        180        1   170
#> [131,]  58   8      0          8        1   130
#> [132,]  54   5      0        180        1   108
#> [133,]  55   3      1          1        1   156
#> [134,]  59   3      1          1        0   172
#> [135,]  57   4      0        180        1   119
#> [136,]  58   6      1          0        1    90
#> [137,]  53  15      1         10        1   130
#> [138,]  55   9      1          2        1   147
#> [139,]  55  13      0        166        1   140
#> [140,]  57   4      1          2        1   185
#> [141,]  53   4      0        147        1   145
#> [142,]  53   7      1          0        1   120
#> [143,]  57  11      1         10        1   129
#> [144,]  54   7      1          0        1   141
#> [145,]  56   4      0          4        0   164
#> [146,]  59  15      1         10        0   140
#> [147,]  58   1      1          1        1   200
#> [148,]  59   9      1          1        1   125
#> [149,]  59   3      0        180        0   120
#> [150,]  60   5      1          1        0   138
#> [151,]  55   5      1          0        1   160
#> [152,]  57  10      1          9        0   103
#> [153,]  59   6      1          0        1   140
#> [154,]  59   5      0        180        1   155
#> [155,]  59   4      1          0        1   152
#> [156,]  58  26      1          0        1   189
#> [157,]  61   9      0          9        1   160
#> [158,]  58   4      1          3        0   120
#> [159,]  60   0      1          0        1    80
#> [160,]  58   8      0        161        1   140
#> [161,]  58  14      1          6        0   190
#> [162,]  61   4      1          3        0   151
#> [163,]  61   9      1          8        0   150
#> [164,]  61   3      1          2        1   102
#> [165,]  58   1      0          1        1   100
#> [166,]  57  13      1         10        0   110
#> [167,]  57   4      1          3        0   138
#> [168,]  57   3      1          2        0   120
#> [169,]  58  19      1         13        1   140
#> [170,]  56  13      1          6        1   158
#> [171,]  56  18      1         11        1   165
#> [172,]  59   9      1          0        1    80
#> [173,]  55   4      1          3        1   160
#> [174,]  58  11      0        172        1   135
#> [175,]  60  12      1          0        1   114
#> [176,]  55   9      1          7        1   135
#> [177,]  61   4      1          0        1   115
#> [178,]  59  11      1          8        1   190
#> [179,]  57   1      0          1        0   126
#> [180,]  57  15      1         13        1   110
#> [181,]  59   5      1          2        0   182
#> [182,]  58   5      1          1        1   135
#> [183,]  59  10      0        180        0   160
#> [184,]  58   8      1          5        0   152
#> [185,]  62  10      1          0        1   153
#> [186,]  62   7      1          2        1   180
#> [187,]  57   3      1          0        0   100
#> [188,]  61  18      0        170        0   140
#> [189,]  61  28      1          7        0   133
#> [190,]  58   8      1          3        1   150
#> [191,]  57   7      0        169        0   180
#> [192,]  60   7      0          7        0   147
#> [193,]  61   6      0          6        0   134
#> [194,]  59  13      1          2        0   198
#> [195,]  57  12      1          9        1   120
#> [196,]  60  17      1          8        1   140
#> [197,]  58   3      1          0        1   146
#> [198,]  62   4      1          3        0   173
#> [199,]  58   2      0         30        0   202
#> [200,]  59  16      1          9        1   133
#> [201,]  63   6      0         28        1   120
#> [202,]  61  13      0         13        0   120
#> [203,]  61   5      0          5        1   110
#> [204,]  57  18      1          9        1    93
#> [205,]  61   5      0          5        1   160
#> [206,]  58  11      1          9        0   179
#> [207,]  57   2      1          1        0   159
#> [208,]  62   1      1          0        1   172
#> [209,]  58   7      0        180        1   150
#> [210,]  63   3      1          1        0   180
#> [211,]  63   1      0        180        1   130
#> [212,]  61   7      0        180        0   135
#> [213,]  62   3      0        180        1   105
#> [214,]  63   4      0        180        1   190
#> [215,]  63  15      1         10        1   126
#> [216,]  64   4      0        180        0   130
#> [217,]  63   4      1          1        0   155
#> [218,]  60  18      1         13        0   132
#> [219,]  59   8      0        180        1   140
#> [220,]  61   9      1          9        1   150
#> [221,]  62   7      0          7        0   150
#> [222,]  59   1      0         22        1   162
#> [223,]  58   2      0        180        0   127
#> [224,]  59   4      0        180        0   196
#> [225,]  60   7      1          5        1   141
#> [226,]  65  13      0        180        1   100
#> [227,]  63   1      0          1        0   162
#> [228,]  63   1      0          1        0   130
#> [229,]  61  15      1         13        0   170
#> [230,]  59   4      0          4        0   149
#> [231,]  60   3      0          3        0   168
#> [232,]  64  10      1          9        0   160
#> [233,]  63  12      1         10        0   200
#> [234,]  59  10      0        180        1   130
#> [235,]  60   8      0         17        1   130
#> [236,]  61   6      1          1        1   117
#> [237,]  64  12      1         11        0   160
#> [238,]  66   1      1          0        1   120
#> [239,]  64   6      1          0        1   140
#> [240,]  63  14      1          9        0   123
#> [241,]  65  36      1         11        0   140
#> [242,]  61  10      1          2        1   194
#> [243,]  63  12      1          9        0   114
#> [244,]  63   7      0        180        0   120
#> [245,]  66   5      1          0        1   110
#> [246,]  65   8      1          0        0   168
#> [247,]  65  10      1          8        1   120
#> [248,]  64   0      0          0        1    90
#> [249,]  64  21      1         10        0   190
#> [250,]  64   9      0        180        0   150
#> [251,]  61   4      0        180        1   113
#> [252,]  63  16      1          7        1   110
#> [253,]  66   6      1          1        1   130
#> [254,]  63  12      0         12        1   150
#> [255,]  62   3      1          1        1   199
#> [256,]  65   6      0          9        0   112
#> [257,]  63   2      1          1        0   180
#> [258,]  67  11      0         11        1   100
#> [259,]  64   2      0          2        0   201
#> [260,]  66  18      1          5        0   142
#> [261,]  66  16      1         11        1   169
#> [262,]  61  14      1          5        0   140
#> [263,]  63   2      1          0        0   140
#> [264,]  64  19      1          8        1   160
#> [265,]  65   8      1          0        1   140
#> [266,]  67   6      0        180        1   170
#> [267,]  65  15      1         11        1   160
#> [268,]  68   5      1          4        1   150
#> [269,]  64  13      1         12        1   150
#> [270,]  64   6      1          0        1   125
#> [271,]  66   7      1          0        1   115
#> [272,]  64  14      1         13        1   150
#> [273,]  65   3      0          3        0   105
#> [274,]  64   0      0          0        1   148
#> [275,]  67   4      1          3        0   130
#> [276,]  66   6      1          0        1   140
#> [277,]  65   2      1          1        1   170
#> [278,]  68   1      0        180        1   166
#> [279,]  67   8      1          1        1   130
#> [280,]  68   5      0          5        1    90
#> [281,]  63  10      0         16        1   160
#> [282,]  66  14      0        180        0   130
#> [283,]  64   1      0          1        1   120
#> [284,]  63   8      1          1        1   162
#> [285,]  65  18      1          3        0   120
#> [286,]  67  11      0         11        0   150
#> [287,]  68  11      0        180        0   160
#> [288,]  66  12      1         10        1   150
#> [289,]  66  11      1          0        0   100
#> [290,]  65   4      1          2        1   145
#> [291,]  69  12      0         15        1   140
#> [292,]  66  15      1         13        1   160
#> [293,]  63   2      0        180        0   150
#> [294,]  69   6      0        180        1   100
#> [295,]  63   8      0        180        1   120
#> [296,]  66   3      0          3        1   138
#> [297,]  69   1      1          0        0   170
#> [298,]  67   1      0        180        1   160
#> [299,]  68  10      1         10        1   150
#> [300,]  65   1      1          0        0   133
#> [301,]  63   2      1          0        0    99
#> [302,]  65  10      1          1        1   148
#> [303,]  66  19      1         12        1   150
#> [304,]  69   6      0         99        1   140
#> [305,]  65   4      1          1        0   130
#> [306,]  64   4      0        179        0   160
#> [307,]  70  15      1         12        1   132
#> [308,]  64  11      0         11        0   125
#> [309,]  64   4      0        180        1   140
#> [310,]  64   0      1          0        1   118
#> [311,]  67   2      0         18        0   131
#> [312,]  66   4      0        180        0   177
#> [313,]  69   4      1          3        1   150
#> [314,]  69  17      1         10        0   140
#> [315,]  69   8      0         93        0   140
#> [316,]  64  21      0         21        1   155
#> [317,]  66   6      0        180        0   140
#> [318,]  68  18      1          0        1   160
#> [319,]  65   6      0        101        1   115
#> [320,]  68   4      0          4        1   190
#> [321,]  71   3      0          5        0   112
#> [322,]  70   7      1          0        1   190
#> [323,]  68   7      0        150        0   210
#> [324,]  71  20      1          0        1   160
#> [325,]  66   9      1          3        1   151
#> [326,]  66   1      1          1        1   165
#> [327,]  69   8      0        180        1   153
#> [328,]  70  14      0        171        0   166
#> [329,]  66   4      0        180        0   130
#> [330,]  67  10      1          9        0   200
#> [331,]  67   6      1          4        0   130
#> [332,]  68  18      1         14        1   170
#> [333,]  69   0      0          0        1   148
#> [334,]  65   2      0        180        0   130
#> [335,]  69   3      1          2        0   151
#> [336,]  67  14      1         13        0   130
#> [337,]  71   7      0          7        0   230
#> [338,]  71   3      0        103        0   133
#> [339,]  69   3      0          3        1   130
#> [340,]  68   6      0        180        0   145
#> [341,]  69   6      1          4        1   174
#> [342,]  72   7      0          7        1   110
#> [343,]  69   8      1          7        1   108
#> [344,]  67   3      0        180        0   110
#> [345,]  69  19      0        180        0   130
#> [346,]  68  18      0         18        1   100
#> [347,]  69  11      1          0        1   120
#> [348,]  66   2      0        180        0   130
#> [349,]  67   7      1          4        0   122
#> [350,]  69   4      1          3        0   132
#> [351,]  68   2      0          7        1   130
#> [352,]  69   8      1          2        0   121
#> [353,]  70   3      0        123        0   130
#> [354,]  70   9      0        180        1   142
#> [355,]  67  22      1          1        1   140
#> [356,]  68   3      0         19        0   135
#> [357,]  67  12      1          8        0   120
#> [358,]  67   1      0          1        1    60
#> [359,]  67   4      0         60        1   136
#> [360,]  69   5      0         76        0   120
#> [361,]  67   8      1          0        1   130
#> [362,]  72  13      1         11        1   195
#> [363,]  68  10      1          8        1   160
#> [364,]  66  24      1         13        0   130
#> [365,]  70  35      1          0        1   105
#> [366,]  70   7      0          7        0   102
#> [367,]  68   7      1          2        0   135
#> [368,]  69  10      1          6        1   120
#> [369,]  72  19      1          8        0   120
#> [370,]  72  12      1         10        0   170
#> [371,]  67   8      0        180        1   170
#> [372,]  70   5      0        180        0   150
#> [373,]  72   2      0          2        1   100
#> [374,]  72   6      1          5        0   115
#> [375,]  71   1      0        173        1   188
#> [376,]  70   3      0        180        0   121
#> [377,]  69   3      0        180        0   220
#> [378,]  71   3      1          2        0   150
#> [379,]  72   5      0         28        0   120
#> [380,]  69  16      1         10        1   140
#> [381,]  68   7      0        180        1   130
#> [382,]  72  16      1          1        1   130
#> [383,]  70   4      0        180        0   180
#> [384,]  73   6      1          0        1   270
#> [385,]  71   2      1          0        1   180
#> [386,]  73   7      0          7        1   140
#> [387,]  68  15      1         13        1   130
#> [388,]  72   6      0        180        1   130
#> [389,]  73   0      0        180        1   161
#> [390,]  74   8      1          0        1    85
#> [391,]  73   4      0        180        1   154
#> [392,]  69   2      1          0        1   110
#> [393,]  71   3      1          1        0   150
#> [394,]  71  15      1         11        0   165
#> [395,]  74  20      0         20        1   180
#> [396,]  68   9      0        180        1   120
#> [397,]  74   0      1          0        1    90
#> [398,]  71  17      1         11        0   160
#> [399,]  71   3      1          2        1   190
#> [400,]  69  12      1          1        1   149
#> [401,]  70  26      1         11        1   120
#> [402,]  74   4      0          4        0   120
#> [403,]  73   4      0         58        1   160
#> [404,]  72   5      1          3        1   160
#> [405,]  70   3      0        180        1   154
#> [406,]  73   6      0        180        0   110
#> [407,]  72  15      1          0        1   150
#> [408,]  71   7      1          2        0   143
#> [409,]  72   8      1          0        1   140
#> [410,]  74   3      0          3        1   150
#> [411,]  73  17      1         11        0   140
#> [412,]  71  14      1         13        1   170
#> [413,]  72  10      1          8        1   153
#> [414,]  69   7      0        180        1   144
#> [415,]  72  15      1         13        0   156
#> [416,]  70   8      0          8        0   120
#> [417,]  71  10      1          9        1   120
#> [418,]  75   2      1          1        0   145
#> [419,]  73  10      1          9        1   146
#> [420,]  73  10      1         10        1   120
#> [421,]  74  15      1          9        1   179
#> [422,]  71   2      0         10        1   112
#> [423,]  73   1      0          1        1    80
#> [424,]  75   9      1          7        0   140
#> [425,]  71  11      1          8        0   110
#> [426,]  71   4      0          4        0   134
#> [427,]  72  15      1         12        1   120
#> [428,]  73  10      1          8        0   120
#> [429,]  70   7      1          4        0   184
#> [430,]  72   1      1          1        0   168
#> [431,]  72   7      0         57        1   145
#> [432,]  73  10      0        180        0   162
#> [433,]  72  11      0         11        1   140
#> [434,]  73  12      1         12        1   140
#> [435,]  75   1      0        180        1   140
#> [436,]  72   4      1          0        1   197
#> [437,]  74  34      1          8        1   233
#> [438,]  76   3      1          0        1   120
#> [439,]  71  32      1         12        1   107
#> [440,]  76   5      0          5        1   130
#> [441,]  77   4      0          4        0   185
#> [442,]  75   3      1          1        0   180
#> [443,]  72   7      1          2        0   142
#> [444,]  71  16      0        180        0   140
#> [445,]  73  10      1         10        0   124
#> [446,]  74   7      0        180        1   150
#> [447,]  74   3      0          3        1   128
#> [448,]  76   1      0        180        0   114
#> [449,]  74  19      1          4        1   200
#> [450,]  75  23      1         14        1   110
#> [451,]  74   2      0        180        0   190
#> [452,]  72   4      0         85        1   120
#> [453,]  72   4      1          3        0   160
#> [454,]  73   4      1          3        1   125
#> [455,]  76  13      1         10        0   110
#> [456,]  75  12      0         12        1   160
#> [457,]  74   8      1          0        1   105
#> [458,]  76  13      1          8        1   148
#> [459,]  75   4      1          2        1   188
#> [460,]  74   6      0        180        0   160
#> [461,]  76   4      0          4        1   155
#> [462,]  73   1      0         52        1   105
#> [463,]  73   0      0        180        0   156
#> [464,]  72   5      0        180        0   120
#> [465,]  76  44      1         10        0   105
#> [466,]  74  10      1          0        1   135
#> [467,]  76   5      1          0        1   167
#> [468,]  74   8      1          8        1   170
#> [469,]  73  33      1         12        1   175
#> [470,]  73  10      1          9        0   146
#> [471,]  77   1      1          0        1    90
#> [472,]  73   7      1          0        0   174
#> [473,]  74   6      0         79        1   140
#> [474,]  74   9      1          8        0   200
#> [475,]  75   6      0        180        0   150
#> [476,]  78  18      0         18        1   144
#> [477,]  77   3      0        180        0   110
#> [478,]  74   2      0        180        0   100
#> [479,]  76  13      1          1        1   170
#> [480,]  78  32      1          9        1   198
#> [481,]  80  10      1          6        1   147
#> [482,]  78   0      0        180        1   212
#> [483,]  78  13      1          5        0   130
#> [484,]  75   5      0        119        1   150
#> [485,]  75  12      1          1        1   120
#> [486,]  75  13      1          6        0   150
#> [487,]  74  10      1          8        0   135
#> [488,]  76   1      0          1        1    83
#> [489,]  79   4      0         80        0   145
#> [490,]  78  12      1          9        0   150
#> [491,]  78  10      0        180        1   130
#> [492,]  76  11      1          0        0   120
#> [493,]  75  11      1          4        0   162
#> [494,]  75   3      0          3        0     0
#> [495,]  76   7      0         29        1   150
#> [496,]  78   6      1          0        1   240
#> [497,]  76   3      1          0        1   140
#> [498,]  78  11      1          1        1   140
#> [499,]  79  11      0        180        0   160
#> [500,]  79   2      1          0        1   121
#> [501,]  78  14      1          0        1   140
#> [502,]  81   1      0          1        0   130
#> [503,]  76   4      0          4        1   160
#> [504,]  76  12      1         10        1   127
#> [505,]  77   6      0          6        1   107
#> [506,]  80   3      1          0        1   120
#> [507,]  76   1      0          1        1   140
#> [508,]  78   7      1          0        1   110
#> [509,]  79   3      0          3        0   120
#> [510,]  77   7      0        180        1   170
#> [511,]  79   4      1          0        1   120
#> [512,]  81   1      0        180        0   120
#> [513,]  77   9      1          4        0   141
#> [514,]  82   5      0          8        1   120
#> [515,]  78   4      0         59        1   112
#> [516,]  80  17      1         12        0   100
#> [517,]  76   7      0        161        0   151
#> [518,]  79  10      0         10        1   120
#> [519,]  81   4      1          2        1   126
#> [520,]  79  28      0        164        0   100
#> [521,]  80   9      0        118        1   186
#> [522,]  80   6      0        173        1   160
#> [523,]  81   3      0        180        0   184
#> [524,]  78   7      0          7        1   147
#> [525,]  78  15      0         15        0   165
#> [526,]  80   5      1          1        1   108
#> [527,]  78   4      0        180        0   175
#> [528,]  76   1      0        166        0   131
#> [529,]  81   4      1          1        1   104
#> [530,]  78  20      1          0        1   109
#> [531,]  80   1      0          1        0   100
#> [532,]  82   3      1          1        1   144
#> [533,]  80   2      1          1        0   168
#> [534,]  79   6      0          6        0   152
#> [535,]  80   6      1          0        1   119
#> [536,]  78   2      0        180        0   148
#> [537,]  82   1      1          0        1    82
#> [538,]  79  10      0        180        1   150
#> [539,]  77   4      0        180        1    98
#> [540,]  81   1      0        108        0   129
#> [541,]  78  12      0        180        0   134
#> [542,]  79   1      0        125        0   193
#> [543,]  82  21      1          2        0   155
#> [544,]  80   6      0          6        1   110
#> [545,]  82   5      0        180        0   110
#> [546,]  83   5      0        180        0   148
#> [547,]  79   7      1          6        0   130
#> [548,]  83   4      0        103        0    97
#> [549,]  81  11      1          8        0   160
#> [550,]  80  11      1          8        0   170
#> [551,]  78  23      1         10        1   145
#> [552,]  79   4      0          4        1   183
#> [553,]  78   9      1          4        1   120
#> [554,]  79   1      0        180        1   170
#> [555,]  81  15      0        180        1   140
#> [556,]  80   7      1          0        1   146
#> [557,]  84   5      1          1        1    85
#> [558,]  83   8      0          8        0   115
#> [559,]  81  16      0         16        1   110
#> [560,]  80   6      1          0        1   150
#> [561,]  80  11      1          8        0   110
#> [562,]  81   8      0        180        0   146
#> [563,]  80   8      1          7        0   160
#> [564,]  79   0      1          0        1    96
#> [565,]  85   4      0        180        0    90
#> [566,]  83   2      0          2        1   155
#> [567,]  82   6      0        128        1   100
#> [568,]  80   3      1          1        1   230
#> [569,]  82  23      1          0        0   110
#> [570,]  84   5      0        180        1   203
#> [571,]  84   4      0          4        1    85
#> [572,]  84   1      0         38        1   205
#> [573,]  83   3      0        180        0   174
#> [574,]  81   4      0         90        1   138
#> [575,]  79   9      1          8        0   150
#> [576,]  85   3      1          2        1   160
#> [577,]  80  13      1          8        1   140
#> [578,]  79   4      0          4        1    60
#> [579,]  80   6      0         71        1   189
#> [580,]  83   1      0          1        1   100
#> [581,]  82  19      0         19        0   120
#> [582,]  80  30      1         13        0   220
#> [583,]  83   9      0        180        0   198
#> [584,]  81  14      1         12        1   128
#> [585,]  83   2      0        154        0   130
#> [586,]  82   0      0          2        1   100
#> [587,]  85   9      1          6        1   160
#> [588,]  83   1      0        180        0   160
#> [589,]  81   4      0          4        0   175
#> [590,]  84  15      1         13        1   110
#> [591,]  81   1      0          1        1   145
#> [592,]  82   5      1          0        1   146
#> [593,]  81   4      0          4        0   160
#> [594,]  86  12      0        180        1   120
#> [595,]  83  12      1          2        1   170
#> [596,]  81  19      1         14        0   120
#> [597,]  82  15      1          0        0   183
#> [598,]  80   2      0         88        0   135
#> [599,]  83   7      0        126        0   135
#> [600,]  86   8      0          8        1   132
#> [601,]  86   3      0          3        1   140
#> [602,]  82   9      0        180        1   134
#> [603,]  81   2      1          0        1   118
#> [604,]  81   4      0        180        0   160
#> [605,]  83   9      0        180        1   149
#> [606,]  83   4      0          4        0   130
#> [607,]  86  12      1          0        1   132
#> [608,]  82  14      1         11        1   103
#> [609,]  84  16      0         70        1   150
#> [610,]  83  19      0         43        0   150
#> [611,]  84   3      1          2        0   125
#> [612,]  84   3      0          3        1   121
#> [613,]  83  13      1         12        0   170
#> [614,]  84   7      1          2        0   148
#> [615,]  87   2      0        180        0   113
#> [616,]  84   9      0         92        1   110
#> [617,]  84   3      0        180        1   170
#> [618,]  86   4      0         38        1   122
#> [619,]  82   4      0          4        0   130
#> [620,]  86  13      0        177        0   163
#> [621,]  85   3      0          3        1   113
#> [622,]  86   6      0          6        1   117
#> [623,]  84  13      0         62        1   100
#> [624,]  88   4      0          4        0   100
#> [625,]  83  20      1          3        1   150
#> [626,]  88   4      0          4        1   115
#> [627,]  85  22      0         22        1   184
#> [628,]  83   9      0         65        1   150
#> [629,]  86   6      0         46        0   173
#> [630,]  83   3      0          3        1   130
#> [631,]  86  15      1          8        1   109
#> [632,]  88   4      0          4        0    86
#> [633,]  89   4      0          4        1   153
#> [634,]  87   1      0          1        0   170
#> [635,]  84   2      0        110        1   174
#> [636,]  87  29      0         29        1    97
#> [637,]  87  15      1          9        1   138
#> [638,]  84   0      0        180        1   136
#> [639,]  90  14      0         14        1   100
#> [640,]  86   4      0        180        1   145
#> [641,]  87   6      1          0        0   125
#> [642,]  90  11      1         10        1   186
#> [643,]  87   6      0        126        1   168
#> [644,]  86   9      1          7        0   130
#> [645,]  91   1      0          1        1    74
#> [646,]  87  43      0        178        1   130
#> [647,]  87   5      0         36        1   150
#> [648,]  90   5      1          0        1   125
#> [649,]  89   3      1          1        1   160
#> [650,]  92   1      0          1        1   167
#> [651,]  88   5      0        158        0   100
#> [652,]  91   5      0        169        1   176
#> [653,]  89  52      0         52        1   130
#> [654,]  92   7      0          7        1   110
#> [655,]  89   4      0          4        1   159
#> [656,]  91   0      0          0        0     0
#> [657,]  89  14      0        180        1    84
#> [658,]  90  18      0        180        0   188
#> [659,]  91   4      1          0        1   120
#> [660,]  90  19      1         11        1   129
#> [661,]  93   8      0        179        1   110
#> [662,]  94   8      0          8        1   142
#> [663,]  92   4      0         76        1   149
#> [664,]  91   1      0        180        0   158
#> [665,]  90   3      0         67        0   162
#> [666,]  95   8      1          5        1   150
#> [667,]  94   3      0         26        1   144
#> [668,]  91  12      0         53        1   212
#> [669,]  93   0      1          0        1   122
#> [670,]  92   2      0          2        0   112
#> 
#> $y
#>   [1]   5.0+   5.0+ 180.0+   2.0+ 115.0  180.0+ 180.0+   5.0+ 180.0+ 180.0+
#>  [11] 180.0+ 180.0+   2.0+ 180.0+   5.0+ 180.0+ 180.0+   3.0  180.0+ 180.0+
#>  [21] 180.0+ 180.0+ 180.0+ 155.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [31] 180.0+ 180.0+ 180.0+ 150.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+   6.0+
#>  [41] 180.0+ 180.0+   5.0+ 161.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [51] 177.0+ 180.0+ 180.0+ 180.0+ 180.0+ 172.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [61]   7.0  180.0+ 180.0+   1.0  180.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+
#>  [71] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0+
#>  [81] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0  180.0+ 180.0+
#>  [91] 180.0+   4.0+  85.0  180.0+ 166.0+  16.0+ 152.0+   7.0+   6.0+ 180.0+
#> [101] 180.0+ 180.0+ 171.0+ 180.0+ 174.0+  28.0    6.0+ 180.0+   9.0+ 180.0+
#> [111]   2.0    7.0+ 180.0+ 180.0+ 180.0+  16.0  180.0+ 180.0+  12.0+ 134.0+
#> [121] 180.0+   8.0    3.0+ 180.0+ 180.0+   1.0  180.0+ 180.0+ 180.0+ 180.0+
#> [131]   8.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  15.0  166.0+   4.0+
#> [141] 147.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+   1.0  180.0+ 180.0+ 180.0+
#> [151] 180.0+ 180.0+  64.0  180.0+ 180.0+ 180.0+   9.0+ 180.0+   0.5  161.0+
#> [161] 171.0+ 180.0+ 180.0+   3.0    1.0  180.0+ 180.0+   3.0+  19.0  180.0+
#> [171] 180.0+   9.0+ 180.0+ 172.0+ 172.0+  24.0  180.0+ 180.0+   1.0+  15.0 
#> [181] 180.0+ 180.0+ 180.0+   8.0+ 180.0+ 180.0+ 180.0+ 170.0   94.0  180.0+
#> [191] 169.0    7.0+   6.0  180.0+ 180.0+ 180.0+   3.0+ 180.0+  30.0  180.0+
#> [201]  28.0   13.0+   5.0   18.0    5.0+ 180.0+ 180.0+   1.0  180.0+ 180.0+
#> [211] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+ 180.0+ 180.0+
#> [221]   7.0+  22.0  180.0+ 180.0+  84.0  180.0+   1.0    1.0  180.0+   4.0+
#> [231]   3.0+ 167.0  180.0+ 180.0+  17.0  180.0+  12.0  180.0+ 180.0+  14.0+
#> [241]  36.0   88.0   12.0  180.0+ 180.0+ 180.0+ 180.0+   0.5  180.0+ 180.0+
#> [251] 180.0+ 180.0+ 180.0+  12.0  180.0+   9.0  180.0+  11.0+   2.0+  18.0+
#> [261] 180.0+ 180.0+   2.0+ 103.0   15.0  180.0+ 180.0+   5.0+  13.0  180.0+
#> [271] 179.0+  14.0+   3.0    0.5+ 180.0+ 180.0+ 175.0+ 180.0+   8.0    5.0 
#> [281]  16.0  180.0+   1.0  180.0+ 123.0+  11.0+ 180.0+  80.0  180.0+   4.0+
#> [291]  15.0  180.0+ 180.0+ 180.0+ 180.0+   3.0  175.0  180.0+  10.0  180.0+
#> [301] 180.0+ 180.0+  19.0+  99.0  180.0+ 179.0+ 180.0+  11.0+ 180.0+   0.5 
#> [311]  18.0  180.0+ 152.0+ 180.0+  93.0   21.0+ 180.0+  18.0+ 101.0    4.0 
#> [321]   5.0    7.0+ 150.0  180.0+ 180.0+   1.0  180.0+ 171.0  180.0+ 174.0+
#> [331]   6.0  180.0+   0.5  180.0+ 180.0+ 180.0+   7.0+ 103.0    3.0+ 180.0+
#> [341]  97.0    7.0    8.0+ 180.0+ 180.0+  18.0  180.0+ 180.0+   7.0  180.0+
#> [351]   7.0    8.0+ 123.0  180.0+  51.0   19.0  180.0+   1.0   60.0   76.0 
#> [361] 180.0+ 132.0   10.0+ 180.0+ 180.0+   7.0+   7.0+ 180.0+ 180.0+  12.0 
#> [371] 180.0+ 180.0+   2.0  180.0+ 173.0+ 180.0+ 180.0+ 180.0+  28.0   16.0+
#> [381] 180.0+  16.0+ 180.0+   6.0  180.0+   7.0+  15.0  180.0+ 180.0+ 180.0+
#> [391] 180.0+   2.0    3.0+ 180.0+  20.0  180.0+   0.5  180.0+   3.0   12.0 
#> [401] 180.0+   4.0+  58.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   3.0 
#> [411] 180.0+  14.0+  10.0+ 180.0+ 180.0+   8.0+ 179.0+ 180.0+ 180.0+  15.0 
#> [421] 180.0+  10.0    1.0  180.0+ 180.0+   4.0+ 180.0+  10.0  104.0+   1.0 
#> [431]  57.0  180.0+  11.0   12.0  180.0+ 180.0+  34.0  180.0+ 177.0+   5.0 
#> [441]   4.0+ 180.0+   7.0  180.0+  10.0  180.0+   3.0  180.0+ 180.0+ 180.0+
#> [451] 180.0+  85.0  180.0+ 180.0+ 174.0+  12.0  180.0+ 180.0+  46.0  180.0+
#> [461]   4.0   52.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+   8.0   33.0  180.0+
#> [471]   1.0    7.0+  79.0  168.0+ 180.0+  18.0  180.0+ 180.0+ 180.0+  32.0 
#> [481]  10.0  180.0+ 172.0  119.0   12.0  180.0+ 180.0+   1.0   80.0  180.0+
#> [491] 180.0+  11.0  152.0+   3.0   29.0  180.0+   3.0+ 180.0+ 180.0+ 180.0+
#> [501] 180.0+   1.0    4.0  180.0+   6.0    3.0+   1.0   43.0    3.0  180.0+
#> [511] 138.0  180.0+  71.0    8.0   59.0   17.0  161.0   10.0+  93.0  164.0 
#> [521] 118.0  173.0  180.0+   7.0+  15.0+   5.0+ 180.0+ 166.0+  71.0   20.0+
#> [531]   1.0  180.0+  10.0    6.0+   6.0  180.0+   1.0  180.0+ 180.0+ 108.0 
#> [541] 180.0+ 125.0  180.0+   6.0  180.0+ 180.0+ 180.0+ 103.0  180.0+ 169.0 
#> [551]  70.0    4.0  180.0+ 180.0+ 180.0+   7.0+ 180.0+   8.0+  16.0  180.0+
#> [561] 180.0+ 180.0+ 180.0+   0.5  180.0+   2.0  128.0    3.0+  62.0  180.0+
#> [571]   4.0   38.0  180.0+  90.0  180.0+ 180.0+ 180.0+   4.0   71.0    1.0 
#> [581]  19.0   30.0  180.0+ 180.0+ 154.0    2.0  180.0+ 180.0+   4.0+ 180.0+
#> [591]   1.0    5.0+   4.0+ 180.0+  77.0  180.0+  83.0   88.0  126.0    8.0 
#> [601]   3.0  180.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+ 174.0   70.0   43.0 
#> [611] 180.0+   3.0   13.0  180.0+ 180.0+  92.0  180.0+  38.0    4.0  177.0 
#> [621]   3.0+   6.0+  62.0    4.0+  20.0    4.0   22.0   65.0   46.0    3.0+
#> [631] 180.0+   4.0    4.0    1.0+ 110.0   29.0  180.0+ 180.0+  14.0  180.0+
#> [641]  25.0   11.0  126.0  180.0+   1.0  178.0+  36.0   89.0    3.0+   1.0 
#> [651] 158.0  169.0   52.0    7.0    4.0    0.5  180.0+ 180.0+   4.0  180.0+
#> [661] 179.0+   8.0+  76.0  180.0+  67.0    8.0   26.0   53.0    0.5    2.0 
#> 
#> $weights
#> NULL
#> 
#> $offset
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Warning: Multiple lambdas have been fit. Lambda will be set to 0.01 (see parameter 's').

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8599005