GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.coxnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.coxnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Caution: This learner is different to learners calling glmnet::cv.glmnet() in that it does not use the internal optimization of parameter lambda. Instead, lambda needs to be tuned by the user (e.g., via mlr3tuning). When lambda is tuned, the glmnet will be trained for each tuning iteration. While fitting the whole path of lambdas would be more efficient, as is done by default in glmnet::glmnet(), tuning/selecting the parameter at prediction time (using parameter s) is currently not supported in mlr3 (at least not in efficient manner). Tuning the s parameter is, therefore, currently discouraged.

When the data are i.i.d. and efficiency is key, we recommend using the respective auto-tuning counterpart in mlr_learners_surv.cv_glmnet(). However, in some situations this is not applicable, usually when data are imbalanced or not i.i.d. (longitudinal, time-series) and tuning requires custom resampling strategies (blocked design, stratification).

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.coxnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exact	logical	FALSE	TRUE, FALSE	-

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	TRUE	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	0.01		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.logistic	character	Newton	Newton, modified.Newton	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvGlmnet

Methods

Public methods

LearnerSurvGlmnet$new()

LearnerSurvGlmnet$selected_features()

LearnerSurvGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.glmnet")
print(learner)
#> 
#> ── <LearnerSurvGlmnet> (surv.glmnet): Regularized Generalized Linear Model ─────
#> • Model: -
#> • Parameters: use_pred_offset=TRUE
#> • Packages: mlr3, mlr3proba, mlr3extralearners, and glmnet
#> • Predict Types: [crank], distr, and lp
#> • Feature Types: logical, integer, and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: offset, selected_features, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#>    Df  %Dev   Lambda
#> 1   0  0.00 0.198800
#> 2   1  0.65 0.181200
#> 3   1  1.19 0.165100
#> 4   2  1.99 0.150400
#> 5   3  3.70 0.137100
#> 6   3  6.65 0.124900
#> 7   3  8.84 0.113800
#> 8   3 10.53 0.103700
#> 9   3 11.88 0.094470
#> 10  3 12.98 0.086080
#> 11  3 13.88 0.078430
#> 12  3 14.64 0.071460
#> 13  3 15.26 0.065110
#> 14  3 15.79 0.059330
#> 15  3 16.23 0.054060
#> 16  3 16.61 0.049260
#> 17  4 16.93 0.044880
#> 18  5 17.22 0.040890
#> 19  6 17.47 0.037260
#> 20  6 17.70 0.033950
#> 21  6 17.90 0.030930
#> 22  6 18.07 0.028190
#> 23  6 18.21 0.025680
#> 24  6 18.33 0.023400
#> 25  6 18.43 0.021320
#> 26  6 18.52 0.019430
#> 27  6 18.59 0.017700
#> 28  6 18.65 0.016130
#> 29  6 18.70 0.014700
#> 30  6 18.74 0.013390
#> 31  6 18.78 0.012200
#> 32  6 18.81 0.011120
#> 33  6 18.83 0.010130
#> 34  6 18.86 0.009230
#> 35  6 18.87 0.008410
#> 36  6 18.89 0.007663
#> 37  6 18.90 0.006982
#> 38  6 18.91 0.006362
#> 39  6 18.92 0.005797
#> 40  6 18.93 0.005282
#> 41  6 18.93 0.004812
#> 42  6 18.94 0.004385
#> 43  6 18.94 0.003995
#> 44  6 18.95 0.003640
#> 45  6 18.95 0.003317
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  28   9      0        180        1   107
#>   [2,]  32   5      1          0        1   121
#>   [3,]  33   2      0          2        0   150
#>   [4,]  35   5      1          2        0   172
#>   [5,]  35  10      1          9        0   106
#>   [6,]  34   5      0          5        0   120
#>   [7,]  35   2      0        180        0   121
#>   [8,]  35   2      1          1        1   112
#>   [9,]  37   9      0        180        1   151
#>  [10,]  38   2      0        115        0   150
#>  [11,]  38  12      1          8        1   120
#>  [12,]  36   5      1          0        1   115
#>  [13,]  33   6      1          1        1   115
#>  [14,]  38  16      1         10        0   160
#>  [15,]  40  12      1          9        0   153
#>  [16,]  37   1      1          0        1   146
#>  [17,]  42   2      0        180        1   100
#>  [18,]  38   5      1          3        0   125
#>  [19,]  42   2      0        180        0   100
#>  [20,]  43   3      1          0        1   100
#>  [21,]  40   1      1          0        1   145
#>  [22,]  42   4      0        180        0   162
#>  [23,]  42  15      1         13        1   125
#>  [24,]  40   3      1          1        0   170
#>  [25,]  42   2      0        180        1   124
#>  [26,]  45   3      0        180        1   154
#>  [27,]  41  10      1          8        0   150
#>  [28,]  44   3      0        180        0   141
#>  [29,]  41  13      1          1        0   140
#>  [30,]  45   6      0        180        1   170
#>  [31,]  41   5      1          4        1   141
#>  [32,]  44   2      1          1        1   150
#>  [33,]  43   2      0        180        1   140
#>  [34,]  45   2      0        180        1   140
#>  [35,]  46   2      1          1        0   126
#>  [36,]  47   4      1          3        0   118
#>  [37,]  48  15      0        180        1   160
#>  [38,]  45   3      0        150        0   130
#>  [39,]  44   3      1          0        1   180
#>  [40,]  46   7      1          2        0   166
#>  [41,]  44   0      1          0        1    96
#>  [42,]  47   4      1          3        1   160
#>  [43,]  49   5      0         73        1   136
#>  [44,]  45   5      0          5        0   141
#>  [45,]  46   2      1          1        1   122
#>  [46,]  46   6      1          0        1   100
#>  [47,]  44   4      1          0        1   114
#>  [48,]  47   2      0        180        0   108
#>  [49,]  45   5      0        180        1   190
#>  [50,]  46   5      1          3        0   130
#>  [51,]  46   4      0        180        1   121
#>  [52,]  46  15      0        180        1   120
#>  [53,]  45   9      1          0        1   145
#>  [54,]  47   3      1          1        1   120
#>  [55,]  48   3      0        180        0   154
#>  [56,]  47   5      1          3        1   130
#>  [57,]  46   3      1          0        1   119
#>  [58,]  49   4      0        180        0   117
#>  [59,]  47  10      0         10        1   140
#>  [60,]  47   7      0        180        0   145
#>  [61,]  50   6      1          2        1   140
#>  [62,]  46   9      1          9        1   122
#>  [63,]  50   7      0        180        1   110
#>  [64,]  49   2      0          2        0   105
#>  [65,]  47   2      0        180        0   150
#>  [66,]  49  23      0        179        1   112
#>  [67,]  52   2      0        180        1   170
#>  [68,]  50   7      1          0        1    92
#>  [69,]  50   4      0          4        1   100
#>  [70,]  51   3      1          2        0   113
#>  [71,]  50   1      1          0        0   150
#>  [72,]  50   9      0        180        0   130
#>  [73,]  47   6      0        180        1   162
#>  [74,]  51   8      0        180        1   140
#>  [75,]  52   2      0        180        0   155
#>  [76,]  46   1      1          1        0   145
#>  [77,]  50   4      1          1        0   150
#>  [78,]  48  17      1         10        0   111
#>  [79,]  49   9      1          3        0   102
#>  [80,]  49  15      0        180        1   160
#>  [81,]  53   5      0         77        0   159
#>  [82,]  53   7      1          0        0   199
#>  [83,]  51   3      1          1        0   140
#>  [84,]  50   2      0          5        1   106
#>  [85,]  50  10      1          6        0   122
#>  [86,]  48   3      1          2        0   150
#>  [87,]  51  25      1          1        0   202
#>  [88,]  49   5      1          2        1   150
#>  [89,]  53   4      0          4        0   140
#>  [90,]  48   6      0        180        0   160
#>  [91,]  51  13      0         99        1   160
#>  [92,]  54   9      1          0        1   138
#>  [93,]  49  16      0         16        0   125
#>  [94,]  54  23      1         10        0   131
#>  [95,]  55   6      1          2        1   114
#>  [96,]  55   4      1          2        0   150
#>  [97,]  54   4      1          0        1   121
#>  [98,]  50   3      0        174        1   153
#>  [99,]  55  28      1         13        1   160
#> [100,]  49   1      0          1        1   110
#> [101,]  50   7      1          1        0   156
#> [102,]  53   9      0          9        1    95
#> [103,]  50   7      1          0        1   127
#> [104,]  55   1      0        180        0   127
#> [105,]  55   2      0          2        0   145
#> [106,]  54   7      1          0        1   100
#> [107,]  56   3      0        180        1   193
#> [108,]  56   2      0        180        0   132
#> [109,]  55   5      1          4        1   120
#> [110,]  52   8      0        180        0   119
#> [111,]  53  18      1          9        1   150
#> [112,]  52  16      0         16        0   152
#> [113,]  52  16      1         14        0   170
#> [114,]  53  15      0         15        1    90
#> [115,]  53   4      0        180        1   150
#> [116,]  55   6      1          5        1   138
#> [117,]  55   2      0        134        1   140
#> [118,]  56   3      0          8        1   139
#> [119,]  55   1      0          2        0   130
#> [120,]  54   2      1          1        0   135
#> [121,]  52   9      1          3        0   170
#> [122,]  54   2      1          1        1   176
#> [123,]  57   5      1          3        1   138
#> [124,]  57   1      0        180        1   156
#> [125,]  52   2      0        180        0   140
#> [126,]  55  11      1          7        0   104
#> [127,]  52  15      1         14        0   130
#> [128,]  53   3      1          0        1   200
#> [129,]  54   5      0        180        1   108
#> [130,]  57   0      0          0        1   150
#> [131,]  53  21      1         13        1   130
#> [132,]  59   3      1          1        0   172
#> [133,]  53  15      1         10        1   130
#> [134,]  56   5      0          5        1   150
#> [135,]  57   4      1          2        1   185
#> [136,]  53   4      0        147        1   145
#> [137,]  55   3      1          2        0   140
#> [138,]  55   5      0          5        1   131
#> [139,]  54   7      1          0        1   141
#> [140,]  59  15      1         10        0   140
#> [141,]  58   9      1          0        1   180
#> [142,]  55   5      1          0        0   140
#> [143,]  56   7      1          5        1   120
#> [144,]  59   9      1          1        1   125
#> [145,]  57   1      0        180        0   148
#> [146,]  60  11      1          9        0   106
#> [147,]  59   3      0        180        0   120
#> [148,]  58   4      1          0        1   160
#> [149,]  57   2      0          2        1   120
#> [150,]  60   5      1          1        0   138
#> [151,]  57   5      0        180        1   130
#> [152,]  55   5      1          0        1   160
#> [153,]  57  10      1          9        0   103
#> [154,]  59   5      0        180        1   155
#> [155,]  58  26      1          0        1   189
#> [156,]  58   4      1          3        0   120
#> [157,]  58   8      0        161        1   140
#> [158,]  58  14      1          6        0   190
#> [159,]  61   4      1          3        0   151
#> [160,]  61   9      1          8        0   150
#> [161,]  61   3      1          2        1   102
#> [162,]  61  20      1         13        0   130
#> [163,]  57  13      1         10        0   110
#> [164,]  57   2      1          0        1   116
#> [165,]  58  10      0         10        1   150
#> [166,]  61   3      0         17        0   143
#> [167,]  56  14      0         45        0   130
#> [168,]  57   3      1          2        0   120
#> [169,]  58  19      1         13        1   140
#> [170,]  56  18      1         11        1   165
#> [171,]  59   9      1          0        1    80
#> [172,]  60  12      1          0        1   114
#> [173,]  55   9      1          7        1   135
#> [174,]  61   4      1          0        1   115
#> [175,]  56   8      1          8        0   120
#> [176,]  61  13      1         12        1   130
#> [177,]  59  11      1          8        1   190
#> [178,]  59   5      1          2        0   182
#> [179,]  58   5      1          1        1   135
#> [180,]  61   8      0         77        0   120
#> [181,]  58   8      1          5        0   152
#> [182,]  62   7      1          2        1   180
#> [183,]  57   3      1          0        0   100
#> [184,]  61  18      0        170        0   140
#> [185,]  61  28      1          7        0   133
#> [186,]  58   8      1          3        1   150
#> [187,]  57   7      0        169        0   180
#> [188,]  61   7      0          7        1   150
#> [189,]  60   7      0          7        0   147
#> [190,]  61   6      0          6        0   134
#> [191,]  59  13      1          2        0   198
#> [192,]  57  12      1          9        1   120
#> [193,]  62   4      1          0        0   160
#> [194,]  60  17      1          8        1   140
#> [195,]  62   4      1          3        0   173
#> [196,]  58   2      0         30        0   202
#> [197,]  59   1      0        180        0   155
#> [198,]  63   6      0         28        1   120
#> [199,]  61  13      0         13        0   120
#> [200,]  61   5      0          5        1   160
#> [201,]  58  11      1          9        0   179
#> [202,]  57   2      1          1        0   159
#> [203,]  62  17      1         10        1   180
#> [204,]  58   7      0        180        1   150
#> [205,]  63   3      1          1        0   180
#> [206,]  61   7      0        180        0   135
#> [207,]  63   4      1          3        0   222
#> [208,]  62   3      0        180        1   105
#> [209,]  63   4      1          1        0   155
#> [210,]  60  18      1         13        0   132
#> [211,]  59   8      0        180        1   140
#> [212,]  61   9      1          9        1   150
#> [213,]  58   9      1          9        0   110
#> [214,]  59   1      0         22        1   162
#> [215,]  58   2      0        180        0   127
#> [216,]  59   4      0        180        0   196
#> [217,]  60   7      0          7        0   140
#> [218,]  60   7      1          1        1    90
#> [219,]  65  13      0        180        1   100
#> [220,]  62   6      0        180        0   170
#> [221,]  61  15      1         13        0   170
#> [222,]  59   4      0          4        0   149
#> [223,]  60   3      0          3        0   168
#> [224,]  64  10      1          9        0   160
#> [225,]  62   6      0          6        0   120
#> [226,]  63  12      1         10        0   200
#> [227,]  60   8      0         17        1   130
#> [228,]  61   6      1          1        1   117
#> [229,]  64  12      1         11        0   160
#> [230,]  66   1      1          0        1   120
#> [231,]  64   6      1          0        1   140
#> [232,]  63  14      1          9        0   123
#> [233,]  63   4      1          3        0   162
#> [234,]  66   3      1          1        0   127
#> [235,]  66   5      1          0        1   110
#> [236,]  65   8      1          0        0   168
#> [237,]  65  10      1          8        1   120
#> [238,]  60   6      0        180        0   130
#> [239,]  64  21      1         10        0   190
#> [240,]  61  12      1         11        0   154
#> [241,]  64   9      0        180        0   150
#> [242,]  61   4      0        180        1   113
#> [243,]  63  16      1          7        1   110
#> [244,]  66   6      1          1        1   130
#> [245,]  62   3      1          1        1   199
#> [246,]  65   6      0          9        0   112
#> [247,]  65   3      1          0        1    80
#> [248,]  63   5      1          4        0   170
#> [249,]  63   2      1          1        0   180
#> [250,]  67  11      0         11        1   100
#> [251,]  66  18      1          5        0   142
#> [252,]  62   9      0        180        0   145
#> [253,]  61  14      1          5        0   140
#> [254,]  63   9      1          8        1   160
#> [255,]  63   3      1          2        0   120
#> [256,]  63   2      1          0        0   140
#> [257,]  65   8      1          0        1   140
#> [258,]  67   6      0        180        1   170
#> [259,]  65  15      1         11        1   160
#> [260,]  68   5      1          4        1   150
#> [261,]  64  13      1         12        1   150
#> [262,]  64   6      1          0        1   125
#> [263,]  66   7      1          0        1   115
#> [264,]  66  13      1          0        0   118
#> [265,]  64  14      1         13        1   150
#> [266,]  65   3      0          3        0   105
#> [267,]  67   4      1          3        0   130
#> [268,]  66   6      1          0        1   140
#> [269,]  65   2      1          1        1   170
#> [270,]  68   1      0        180        1   166
#> [271,]  64  10      1          9        1   110
#> [272,]  67   8      1          1        1   130
#> [273,]  68   5      0          5        1    90
#> [274,]  63  10      0         16        1   160
#> [275,]  66  14      0        180        0   130
#> [276,]  65  17      1         14        1   100
#> [277,]  65  18      1          3        0   120
#> [278,]  63  10      0         18        1   130
#> [279,]  67  11      0         11        0   150
#> [280,]  68  11      0        180        0   160
#> [281,]  68  14      0         79        0   172
#> [282,]  66  12      1         10        1   150
#> [283,]  65  15      1         12        1   150
#> [284,]  66  11      1          0        0   100
#> [285,]  65   4      1          2        1   145
#> [286,]  69  12      0         15        1   140
#> [287,]  66  15      1         13        1   160
#> [288,]  63   2      0        180        0   150
#> [289,]  65  11      1          6        0   130
#> [290,]  69  21      1         10        0   180
#> [291,]  69   6      0        180        1   100
#> [292,]  66   9      1          8        0   130
#> [293,]  68  14      1         13        1   140
#> [294,]  65   8      1          0        1    90
#> [295,]  68  10      1         10        1   150
#> [296,]  65   1      1          0        0   133
#> [297,]  67   7      1          4        1   130
#> [298,]  65   6      0          6        0    80
#> [299,]  65  10      1          1        1   148
#> [300,]  66  19      1         12        1   150
#> [301,]  67  12      1         12        0   160
#> [302,]  69   6      0         99        1   140
#> [303,]  65   4      1          1        0   130
#> [304,]  64   4      0        179        0   160
#> [305,]  66   4      0        180        1   130
#> [306,]  70  15      1         12        1   132
#> [307,]  64  11      0         11        0   125
#> [308,]  64   4      0        180        1   140
#> [309,]  64   0      1          0        1   118
#> [310,]  67   2      0         18        0   131
#> [311,]  66   7      1          5        1   131
#> [312,]  66   4      0        180        0   177
#> [313,]  68   4      1          0        1   160
#> [314,]  69   4      1          3        1   150
#> [315,]  69   8      0         93        0   140
#> [316,]  64  21      0         21        1   155
#> [317,]  66   6      0        180        0   140
#> [318,]  65   1      0          1        1   120
#> [319,]  68  18      1          0        1   160
#> [320,]  65   6      0        101        1   115
#> [321,]  71   3      0          5        0   112
#> [322,]  70   7      1          0        1   190
#> [323,]  68   7      0        150        0   210
#> [324,]  66   9      1          3        1   151
#> [325,]  70   4      1          0        1   180
#> [326,]  66   4      0        180        0   130
#> [327,]  67  10      1          9        0   200
#> [328,]  67   6      1          4        0   130
#> [329,]  69   0      0          0        1   148
#> [330,]  69   3      1          2        0   151
#> [331,]  67  14      1         13        0   130
#> [332,]  69   8      0        180        1   180
#> [333,]  71   7      0          7        0   230
#> [334,]  66   2      0          2        1   228
#> [335,]  71   6      0         45        1   158
#> [336,]  69   5      0          5        1   142
#> [337,]  71   3      0        103        0   133
#> [338,]  67   5      0          5        0   130
#> [339,]  69   8      1          5        1   195
#> [340,]  72   3      1          0        1   132
#> [341,]  72   7      0          7        1   110
#> [342,]  69   8      1          7        1   108
#> [343,]  66   2      1          1        0   123
#> [344,]  69  19      0        180        0   130
#> [345,]  66   2      0        180        0   130
#> [346,]  67   7      1          4        0   122
#> [347,]  69   4      1          3        0   132
#> [348,]  68   2      0          7        1   130
#> [349,]  69   8      1          2        0   121
#> [350,]  67  13      1          9        0   130
#> [351,]  70   3      0        123        0   130
#> [352,]  70   9      0        180        1   142
#> [353,]  67  22      1          1        1   140
#> [354,]  68   3      0         19        0   135
#> [355,]  69   1      0          1        1   110
#> [356,]  67   1      0          1        1    60
#> [357,]  67   4      0         60        1   136
#> [358,]  67   8      1          0        1   130
#> [359,]  72  13      1         11        1   195
#> [360,]  68  10      1          8        1   160
#> [361,]  72  30      1          0        1   145
#> [362,]  70   7      0          7        0   102
#> [363,]  68   7      1          2        0   135
#> [364,]  73  20      1          0        1   170
#> [365,]  71   6      0          9        0   120
#> [366,]  69  10      1          6        1   120
#> [367,]  70  11      0        180        1   210
#> [368,]  72  19      1          8        0   120
#> [369,]  67   9      0        180        0   158
#> [370,]  73  13      0        152        1   130
#> [371,]  70   5      0        180        0   150
#> [372,]  72   2      0          2        1   100
#> [373,]  72   6      1          5        0   115
#> [374,]  71   1      0        173        1   188
#> [375,]  68  23      0        180        1   220
#> [376,]  70   3      0        180        0   121
#> [377,]  69   3      0        180        0   220
#> [378,]  71   3      1          2        0   150
#> [379,]  68   4      1          3        0   210
#> [380,]  72   5      0         28        0   120
#> [381,]  69  16      1         10        1   140
#> [382,]  69   8      1          1        0   164
#> [383,]  68   7      0        180        1   130
#> [384,]  72  16      1          1        1   130
#> [385,]  70   4      0        180        0   180
#> [386,]  69   1      1          0        0   155
#> [387,]  73   6      1          0        1   270
#> [388,]  72   8      1          1        1   150
#> [389,]  71   2      1          0        1   180
#> [390,]  68  15      1         13        1   130
#> [391,]  70   3      0          3        1   159
#> [392,]  73   0      0        180        1   161
#> [393,]  69   2      1          0        1   110
#> [394,]  71   3      1          1        0   150
#> [395,]  74  20      0         20        1   180
#> [396,]  68   9      0        180        1   120
#> [397,]  71  20      1         10        0   140
#> [398,]  74   0      1          0        1    90
#> [399,]  73   3      1          0        1   136
#> [400,]  71  17      1         11        0   160
#> [401,]  71   8      1          7        0   149
#> [402,]  71   3      1          2        1   190
#> [403,]  69  12      1          1        1   149
#> [404,]  70  26      1         11        1   120
#> [405,]  74   4      0          4        0   120
#> [406,]  73   4      0         58        1   160
#> [407,]  72   5      1          3        1   160
#> [408,]  73   6      0        180        0   110
#> [409,]  72  15      1          0        1   150
#> [410,]  72   8      1          0        1   140
#> [411,]  74   3      0          3        1   150
#> [412,]  73  17      1         11        0   140
#> [413,]  69   2      1          1        1    80
#> [414,]  71  14      1         13        1   170
#> [415,]  72  10      1          8        1   153
#> [416,]  69   7      0        180        1   144
#> [417,]  72  15      1         13        0   156
#> [418,]  70   8      0          8        0   120
#> [419,]  75   1      0          1        0   133
#> [420,]  75   2      1          1        0   145
#> [421,]  73  10      1          9        1   146
#> [422,]  73  10      1         10        1   120
#> [423,]  74  15      1          9        1   179
#> [424,]  71   2      0         10        1   112
#> [425,]  75   9      1          7        0   140
#> [426,]  71  11      1          8        0   110
#> [427,]  71   4      0          4        0   134
#> [428,]  72  15      1         12        1   120
#> [429,]  73  10      1          8        0   120
#> [430,]  70   7      1          4        0   184
#> [431,]  72   1      1          1        0   168
#> [432,]  72   7      0         57        1   145
#> [433,]  72  11      0         11        1   140
#> [434,]  70   3      0          3        0   150
#> [435,]  73   5      1          3        1   112
#> [436,]  72   4      1          0        1   197
#> [437,]  71   3      1          0        0   144
#> [438,]  73   5      0        180        0   126
#> [439,]  73   4      0        180        0   124
#> [440,]  74  34      1          8        1   233
#> [441,]  76   3      1          0        1   120
#> [442,]  71  32      1         12        1   107
#> [443,]  72   5      0        180        0   154
#> [444,]  72   3      0        180        0   160
#> [445,]  76   5      0          5        1   130
#> [446,]  77   4      0          4        0   185
#> [447,]  73  15      0         15        1   160
#> [448,]  71  16      0        180        0   140
#> [449,]  74   3      0          3        1   128
#> [450,]  76   1      0        180        0   114
#> [451,]  76   8      1          0        1   141
#> [452,]  74   2      0        180        0   190
#> [453,]  76  17      1          0        1   200
#> [454,]  73   4      1          3        1   125
#> [455,]  76  13      1         10        0   110
#> [456,]  75   4      1          0        1   122
#> [457,]  75   7      0          7        0   190
#> [458,]  75   0      0          0        1   130
#> [459,]  73  13      1         11        0   195
#> [460,]  74   8      1          0        1   105
#> [461,]  76   4      0          4        1   155
#> [462,]  73   1      0         52        1   105
#> [463,]  73   0      0        180        0   156
#> [464,]  72   5      0        180        0   120
#> [465,]  76  44      1         10        0   105
#> [466,]  76   5      0        180        0   185
#> [467,]  74  10      1          0        1   135
#> [468,]  76   5      1          0        1   167
#> [469,]  74   8      1          8        1   170
#> [470,]  73  33      1         12        1   175
#> [471,]  77   5      1          0        0   123
#> [472,]  77  12      1          9        1   100
#> [473,]  77   1      1          0        1    90
#> [474,]  73   7      1          0        0   174
#> [475,]  74   6      0         79        1   140
#> [476,]  74   9      1          8        0   200
#> [477,]  75   6      0        180        0   150
#> [478,]  74   2      1          0        1   130
#> [479,]  77   3      0        180        0   110
#> [480,]  73  11      1          2        1   110
#> [481,]  78   7      0          7        1   133
#> [482,]  74  15      0        180        1   172
#> [483,]  78   8      1          6        1   110
#> [484,]  74   7      0          7        0   161
#> [485,]  76  13      1          1        1   170
#> [486,]  78  32      1          9        1   198
#> [487,]  80  10      1          6        1   147
#> [488,]  78   0      0        180        1   212
#> [489,]  78  13      1          5        0   130
#> [490,]  75   5      0        119        1   150
#> [491,]  75  12      1          1        1   120
#> [492,]  78  15      0        180        1   270
#> [493,]  74  10      1          8        0   135
#> [494,]  76   1      0          1        1    83
#> [495,]  79   4      0         80        0   145
#> [496,]  78  12      1          9        0   150
#> [497,]  78   2      1          1        0   130
#> [498,]  77   2      1          0        1   143
#> [499,]  78  10      0        180        1   130
#> [500,]  75  11      1          4        0   162
#> [501,]  75   3      0          3        0     0
#> [502,]  76   7      0         29        1   150
#> [503,]  77  24      0         24        1   160
#> [504,]  79   8      0         32        1   120
#> [505,]  80   9      0         23        1   128
#> [506,]  80   6      0          6        1   150
#> [507,]  76   3      1          0        1   140
#> [508,]  79  11      0        180        0   160
#> [509,]  79   2      1          0        1   121
#> [510,]  78  14      1          0        1   140
#> [511,]  81   1      0          1        0   130
#> [512,]  78  11      1          8        1   118
#> [513,]  79   4      0          4        1   125
#> [514,]  76  12      1         10        1   127
#> [515,]  77   6      0          6        1   107
#> [516,]  80   3      1          0        1   120
#> [517,]  75   2      1          1        1   204
#> [518,]  76   1      0          1        1   140
#> [519,]  78   7      1          0        1   110
#> [520,]  79   3      0          3        0   120
#> [521,]  77   7      0        180        1   170
#> [522,]  77   6      0          6        1   144
#> [523,]  79   4      1          0        1   120
#> [524,]  80  15      1         12        1   150
#> [525,]  77   9      1          4        0   141
#> [526,]  80  17      1         12        0   100
#> [527,]  80  15      1          0        1    90
#> [528,]  81   4      1          2        1   126
#> [529,]  79   1      0         37        1   140
#> [530,]  81   3      0        180        0   184
#> [531,]  81   2      0        175        0   172
#> [532,]  78   4      0        180        0   175
#> [533,]  79   3      0          3        1   101
#> [534,]  76   1      0        166        0   131
#> [535,]  78  20      1          0        1   109
#> [536,]  80   1      0          1        0   100
#> [537,]  77  10      1          8        1   130
#> [538,]  79   6      0          6        0   152
#> [539,]  78   2      0        180        0   148
#> [540,]  80   5      0          5        1   130
#> [541,]  82   1      1          0        1    82
#> [542,]  79  10      0        180        1   150
#> [543,]  81   1      0        108        0   129
#> [544,]  79   1      0        125        0   193
#> [545,]  82  21      1          2        0   155
#> [546,]  84  22      1         10        0   180
#> [547,]  79   4      0          4        1   121
#> [548,]  80   6      0          6        1   110
#> [549,]  82   5      0        180        0   110
#> [550,]  83   5      0        180        0   148
#> [551,]  83   4      0        103        0    97
#> [552,]  80  11      1          8        0   170
#> [553,]  78  23      1         10        1   145
#> [554,]  79   4      0          4        1   183
#> [555,]  82   8      1          1        0   128
#> [556,]  79   1      0        180        1   170
#> [557,]  81  15      0        180        1   140
#> [558,]  80   7      1          0        1   146
#> [559,]  84   5      1          1        1    85
#> [560,]  81  20      1          9        0   170
#> [561,]  83   8      0          8        0   115
#> [562,]  81  16      0         16        1   110
#> [563,]  80   6      1          0        1   150
#> [564,]  79   7      0        177        0   197
#> [565,]  85   4      0        180        0    90
#> [566,]  82   6      0        128        1   100
#> [567,]  84   4      0        167        0   198
#> [568,]  80   3      1          1        1   230
#> [569,]  84   4      0          4        1    85
#> [570,]  81   1      0          1        1   150
#> [571,]  79   9      1          8        0   150
#> [572,]  85   3      1          2        1   160
#> [573,]  84   4      0         89        1   129
#> [574,]  80   2      1          0        1   130
#> [575,]  80   6      0         71        1   189
#> [576,]  82  19      0         19        0   120
#> [577,]  80  30      1         13        0   220
#> [578,]  83   9      0        180        0   198
#> [579,]  83   3      0        114        0    98
#> [580,]  81  14      1         12        1   128
#> [581,]  83   2      0        154        0   130
#> [582,]  83   1      0        180        0   160
#> [583,]  81   4      0          4        0   175
#> [584,]  84  15      1         13        1   110
#> [585,]  81   1      0          1        1   145
#> [586,]  81  12      0         12        1   163
#> [587,]  82   5      1          0        1   146
#> [588,]  81   4      0          4        0   160
#> [589,]  86  12      0        180        1   120
#> [590,]  81  19      1         14        0   120
#> [591,]  80   2      0         88        0   135
#> [592,]  84   6      0        165        0   145
#> [593,]  86   3      0          3        1   140
#> [594,]  85   3      0          3        1   118
#> [595,]  81   2      1          0        1   118
#> [596,]  81   4      0        180        0   160
#> [597,]  83   9      0        180        1   149
#> [598,]  83   4      0          4        0   130
#> [599,]  86  12      1          0        1   132
#> [600,]  86   6      1          0        1   140
#> [601,]  84  16      0         70        1   150
#> [602,]  83  19      0         43        0   150
#> [603,]  84   3      1          2        0   125
#> [604,]  86   2      0        180        1   169
#> [605,]  88  14      1          3        1   130
#> [606,]  87   2      0        180        0   113
#> [607,]  84   9      0         92        1   110
#> [608,]  84   3      0        180        1   170
#> [609,]  86   4      0         38        1   122
#> [610,]  82   4      0          4        0   130
#> [611,]  86  13      0        177        0   163
#> [612,]  86   6      0          6        1   117
#> [613,]  84  13      0         62        1   100
#> [614,]  86   6      1          1        0   112
#> [615,]  83  20      1          3        1   150
#> [616,]  88   4      0          4        1   115
#> [617,]  86   9      1          7        1   142
#> [618,]  87   2      0        180        1   130
#> [619,]  86   6      0         46        0   173
#> [620,]  88   3      0        115        0   110
#> [621,]  88   2      0        180        1    68
#> [622,]  83   3      0          3        1   130
#> [623,]  86  15      1          8        1   109
#> [624,]  88   4      0          4        0    86
#> [625,]  89   5      0        119        1   140
#> [626,]  87   6      0        180        1   110
#> [627,]  85   8      0          8        1   136
#> [628,]  89  10      0         46        1   170
#> [629,]  90  14      0         14        1   100
#> [630,]  88   1      0          1        0   135
#> [631,]  86   4      0        180        1   145
#> [632,]  91   8      0          8        0   100
#> [633,]  87   6      1          0        0   125
#> [634,]  91  10      0        145        0   135
#> [635,]  88   7      0         24        0   119
#> [636,]  88   8      0         50        1   154
#> [637,]  90  11      1         10        1   186
#> [638,]  87   6      0        126        1   168
#> [639,]  86  10      0        180        1   137
#> [640,]  91   1      0          1        1    74
#> [641,]  87  43      0        178        1   130
#> [642,]  87   5      0         36        1   150
#> [643,]  90   5      1          0        1   125
#> [644,]  88   3      1          2        0   159
#> [645,]  89   3      1          1        1   160
#> [646,]  92   1      0          1        1   167
#> [647,]  91   3      0         33        1   137
#> [648,]  87   7      0         74        1   105
#> [649,]  89  12      1          0        1   130
#> [650,]  89   2      0        168        0   118
#> [651,]  91   5      0        169        1   176
#> [652,]  89  52      0         52        1   130
#> [653,]  92   7      0          7        1   110
#> [654,]  89   4      0          4        1   159
#> [655,]  89  14      0        180        1    84
#> [656,]  90  18      0        180        0   188
#> [657,]  91   4      1          0        1   120
#> [658,]  94   6      0         50        0    78
#> [659,]  91   2      0          2        1   116
#> [660,]  93   8      0        179        1   110
#> [661,]  94   8      0          8        1   142
#> [662,]  92   4      0         76        1   149
#> [663,]  91   1      0        180        0   158
#> [664,]  90  16      0         16        1   106
#> [665,]  95   8      1          5        1   150
#> [666,]  94   3      0         26        1   144
#> [667,]  93   0      1          0        1   122
#> [668,]  92   2      0          2        0   112
#> [669,]  93   4      0        180        1   135
#> [670,]  96   3      1          0        1   104
#> 
#> $y
#>   [1] 180.0+   5.0+   2.0+   5.0+ 180.0+   5.0+ 180.0+   2.0+ 180.0+ 115.0 
#>  [11]  12.0    5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0+ 180.0+   3.0 
#>  [21] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [31]   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 150.0  180.0+ 180.0+
#>  [41] 180.0+ 180.0+  73.0    5.0+ 161.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0+
#>  [51] 180.0+ 180.0+ 177.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  10.0+ 180.0+
#>  [61] 180.0+ 180.0+ 180.0+   2.0  180.0+ 179.0+ 180.0+ 180.0+   4.0+ 180.0+
#>  [71] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  88.0+ 180.0+ 180.0+
#>  [81]  77.0  180.0+ 180.0+   5.0  180.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+
#>  [91]  99.0  180.0+  16.0+ 152.0+   6.0+ 180.0+ 180.0+ 174.0+  28.0    1.0 
#> [101] 180.0+   9.0+ 180.0+ 180.0+   2.0    7.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [111] 180.0+  16.0+  16.0   15.0+ 180.0+ 180.0+ 134.0+   8.0    2.0  180.0+
#> [121] 180.0+ 180.0+ 140.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   0.5 
#> [131] 180.0+ 180.0+ 180.0+   5.0+   4.0+ 147.0+ 180.0+   5.0+ 180.0+ 180.0+
#> [141]   9.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0  180.0+
#> [151] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 161.0+ 171.0+ 180.0+ 180.0+
#> [161]   3.0  180.0+ 180.0+ 180.0+  10.0+  17.0   45.0    3.0+  19.0  180.0+
#> [171]   9.0+ 172.0+  24.0  180.0+   8.0  180.0+ 180.0+ 180.0+ 180.0+  77.0 
#> [181]   8.0+ 180.0+ 180.0+ 170.0   94.0  180.0+ 169.0    7.0    7.0+   6.0 
#> [191] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  30.0  180.0+  28.0   13.0+   5.0+
#> [201] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+
#> [211] 180.0+ 180.0+   9.0   22.0  180.0+ 180.0+   7.0+ 180.0+ 180.0+ 180.0+
#> [221] 180.0+   4.0+   3.0+ 167.0    6.0+ 180.0+  17.0  180.0+  12.0  180.0+
#> [231] 180.0+  14.0+ 180.0+   3.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  12.0+
#> [241] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   9.0    3.0  180.0+ 180.0+  11.0+
#> [251]  18.0+ 180.0+ 180.0+ 180.0+   3.0+   2.0+  15.0  180.0+ 180.0+   5.0+
#> [261]  13.0  180.0+ 179.0+ 166.0+  14.0+   3.0  180.0+ 180.0+ 175.0+ 180.0+
#> [271] 180.0+   8.0    5.0   16.0  180.0+ 180.0+ 123.0+  18.0   11.0+ 180.0+
#> [281]  79.0   80.0   15.0+ 180.0+   4.0+  15.0  180.0+ 180.0+ 180.0+ 174.0+
#> [291] 180.0+ 180.0+ 180.0+   8.0+  10.0  180.0+ 180.0+   6.0  180.0+  19.0+
#> [301]  12.0   99.0  180.0+ 179.0+ 180.0+ 180.0+  11.0+ 180.0+   0.5   18.0 
#> [311]   7.0+ 180.0+ 180.0+ 152.0+  93.0   21.0+ 180.0+   1.0   18.0+ 101.0 
#> [321]   5.0    7.0+ 150.0  180.0+ 180.0+ 180.0+ 174.0+   6.0    0.5  180.0+
#> [331] 180.0+ 180.0+   7.0+   2.0   45.0    5.0+ 103.0    5.0+ 180.0+ 180.0+
#> [341]   7.0    8.0+   2.0+ 180.0+ 180.0+   7.0  180.0+   7.0    8.0+  13.0+
#> [351] 123.0  180.0+  51.0   19.0    1.0    1.0   60.0  180.0+ 132.0   10.0+
#> [361] 162.0    7.0+   7.0+ 124.0    9.0  180.0+ 180.0+ 180.0+ 180.0+ 152.0 
#> [371] 180.0+   2.0  180.0+ 173.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  28.0 
#> [381]  16.0+ 180.0+ 180.0+  16.0+ 180.0+ 180.0+   6.0  180.0+ 180.0+  15.0 
#> [391]   3.0+ 180.0+   2.0    3.0+  20.0  180.0+  20.0    0.5  180.0+ 180.0+
#> [401]   8.0    3.0   12.0  180.0+   4.0+  58.0  180.0+ 180.0+ 180.0+ 180.0+
#> [411]   3.0  180.0+   2.0   14.0+  10.0+ 180.0+ 180.0+   8.0+   1.0  180.0+
#> [421] 180.0+  15.0  180.0+  10.0  180.0+ 180.0+   4.0+ 180.0+  10.0  104.0+
#> [431]   1.0   57.0   11.0    3.0+   5.0  180.0+ 180.0+ 180.0+ 180.0+  34.0 
#> [441] 180.0+ 177.0+ 180.0+ 180.0+   5.0    4.0+  15.0+ 180.0+   3.0  180.0+
#> [451] 180.0+ 180.0+  17.0+ 180.0+ 174.0+   4.0    7.0    0.5  180.0+ 180.0+
#> [461]   4.0   52.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   8.0   33.0 
#> [471]   5.0  180.0+   1.0    7.0+  79.0  168.0+ 180.0+ 176.0+ 180.0+  11.0 
#> [481]   7.0  180.0+   8.0+   7.0  180.0+  32.0   10.0  180.0+ 172.0  119.0 
#> [491]  12.0  180.0+ 180.0+   1.0   80.0  180.0+ 180.0+   2.0  180.0+ 152.0+
#> [501]   3.0   29.0   24.0   32.0   23.0    6.0    3.0+ 180.0+ 180.0+ 180.0+
#> [511]   1.0   11.0    4.0  180.0+   6.0    3.0+   2.0+   1.0   43.0    3.0 
#> [521] 180.0+   6.0  138.0  180.0+  71.0   17.0  180.0+  93.0   37.0  180.0+
#> [531] 175.0+ 180.0+   3.0  166.0+  20.0+   1.0   10.0    6.0+ 180.0+   5.0 
#> [541]   1.0  180.0+ 108.0  125.0  180.0+ 180.0+   4.0    6.0  180.0+ 180.0+
#> [551] 103.0  169.0   70.0    4.0  180.0+ 180.0+ 180.0+   7.0+ 180.0+  20.0 
#> [561]   8.0+  16.0  180.0+ 177.0+ 180.0+ 128.0  167.0    3.0+   4.0    1.0 
#> [571] 180.0+ 180.0+  89.0  180.0+  71.0   19.0   30.0  180.0+ 114.0  180.0+
#> [581] 154.0  180.0+   4.0+ 180.0+   1.0   12.0    5.0+   4.0+ 180.0+ 180.0+
#> [591]  88.0  165.0    3.0    3.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+   6.0 
#> [601]  70.0   43.0  180.0+ 180.0+  14.0  180.0+  92.0  180.0+  38.0    4.0 
#> [611] 177.0    6.0+  62.0    6.0+  20.0    4.0   11.0  180.0+  46.0  115.0 
#> [621] 180.0+   3.0+ 180.0+   4.0  119.0  180.0+   8.0   46.0   14.0    1.0+
#> [631] 180.0+   8.0   25.0  145.0   24.0   50.0   11.0  126.0  180.0+   1.0 
#> [641] 178.0+  36.0   89.0   75.0    3.0+   1.0   33.0   74.0  180.0+ 168.0 
#> [651] 169.0   52.0    7.0    4.0  180.0+ 180.0+   4.0   50.0    2.0  179.0+
#> [661]   8.0+  76.0  180.0+  16.0    8.0   26.0    0.5    2.0  180.0+   3.0 
#> 
#> $weights
#> NULL
#> 
#> $offset
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Warning: Multiple lambdas have been fit. Lambda will be set to 0.01 (see parameter 's').

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8355353