GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::glmnet() from package glmnet.

Details

This learner returns two prediction types:

lp: a vector of linear predictors (relative risk scores), one per observation. Calculated using glmnet::predict.coxnet().

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.coxnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used.

Caution: This learner is different to learners calling glmnet::cv.glmnet() in that it does not use the internal optimization of parameter lambda. Instead, lambda needs to be tuned by the user (e.g., via mlr3tuning). When lambda is tuned, the glmnet will be trained for each tuning iteration. While fitting the whole path of lambdas would be more efficient, as is done by default in glmnet::glmnet(), tuning/selecting the parameter at prediction time (using parameter s) is currently not supported in mlr3 (at least not in efficient manner). Tuning the s parameter is, therefore, currently discouraged.

When the data are i.i.d. and efficiency is key, we recommend using the respective auto-tuning counterpart in mlr_learners_surv.cv_glmnet(). However, in some situations this is not applicable, usually when data are imbalanced or not i.i.d. (longitudinal, time-series) and tuning requires custom resampling strategies (blocked design, stratification).

Custom mlr3 parameters

family is set to "cox" and cannot be changed.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exact	logical	FALSE	TRUE, FALSE	-

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

newoffset	untyped	-		-

nlambda	integer	100		\([1, \infty)\)

offset	untyped	NULL		-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	0.01		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.logistic	character	Newton	Newton, modified.Newton	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvGlmnet

Methods

Public methods

LearnerSurvGlmnet$new()

LearnerSurvGlmnet$selected_features()

LearnerSurvGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.glmnet")
print(learner)
#> <LearnerSurvGlmnet:surv.glmnet>: Regularized Generalized Linear Model
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3proba, mlr3extralearners, glmnet
#> * Predict Types:  [crank], distr, lp
#> * Feature Types: logical, integer, numeric
#> * Properties: selected_features, weights

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#>    Df  %Dev   Lambda
#> 1   0  0.00 0.191800
#> 2   1  0.62 0.174800
#> 3   2  1.34 0.159200
#> 4   2  2.20 0.145100
#> 5   2  2.94 0.132200
#> 6   3  4.74 0.120500
#> 7   3  7.18 0.109800
#> 8   3  8.99 0.100000
#> 9   3 10.40 0.091120
#> 10  3 11.52 0.083030
#> 11  3 12.42 0.075650
#> 12  3 13.17 0.068930
#> 13  3 13.78 0.062810
#> 14  3 14.30 0.057230
#> 15  4 14.75 0.052140
#> 16  4 15.13 0.047510
#> 17  5 15.47 0.043290
#> 18  6 15.75 0.039440
#> 19  6 16.01 0.035940
#> 20  6 16.23 0.032750
#> 21  6 16.42 0.029840
#> 22  6 16.57 0.027190
#> 23  6 16.70 0.024770
#> 24  6 16.81 0.022570
#> 25  6 16.91 0.020570
#> 26  6 16.99 0.018740
#> 27  6 17.05 0.017070
#> 28  6 17.11 0.015560
#> 29  6 17.15 0.014180
#> 30  6 17.19 0.012920
#> 31  6 17.22 0.011770
#> 32  6 17.25 0.010720
#> 33  6 17.28 0.009771
#> 34  6 17.29 0.008903
#> 35  6 17.31 0.008112
#> 36  6 17.32 0.007391
#> 37  6 17.33 0.006734
#> 38  6 17.34 0.006136
#> 39  6 17.35 0.005591
#> 40  6 17.36 0.005094
#> 41  6 17.36 0.004642
#> 42  6 17.37 0.004229
#> 43  6 17.37 0.003854
#> 44  6 17.38 0.003511
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  28   9      0        180        1   107
#>   [2,]  35   5      1          2        0   172
#>   [3,]  35  10      1          9        0   106
#>   [4,]  34   5      0          5        0   120
#>   [5,]  35   2      1          1        1   112
#>   [6,]  37   9      0        180        1   151
#>   [7,]  38  13      1          0        1   161
#>   [8,]  38   2      0        115        0   150
#>   [9,]  35   0      0        180        1   119
#>  [10,]  38  12      1          8        1   120
#>  [11,]  36   5      1          0        1   115
#>  [12,]  33   6      1          1        1   115
#>  [13,]  38  16      1         10        0   160
#>  [14,]  38  12      1         11        1    92
#>  [15,]  40  12      1          9        0   153
#>  [16,]  37   1      1          0        1   146
#>  [17,]  40   2      1          1        1   148
#>  [18,]  42   2      0        180        1   100
#>  [19,]  42   2      0        180        0   100
#>  [20,]  43   3      1          0        1   100
#>  [21,]  41   2      1          1        0   166
#>  [22,]  40   1      1          0        1   145
#>  [23,]  43   4      1          0        1   130
#>  [24,]  42   4      0        180        0   162
#>  [25,]  42  15      1         13        1   125
#>  [26,]  40   3      1          1        0   170
#>  [27,]  42  12      1         10        1   170
#>  [28,]  43   2      1          1        1   116
#>  [29,]  42   2      0        180        1   124
#>  [30,]  44   5      1          1        0   170
#>  [31,]  41  10      1          8        0   150
#>  [32,]  44   3      0        180        0   141
#>  [33,]  41  13      1          1        0   140
#>  [34,]  45   9      1          7        0   110
#>  [35,]  41   5      1          4        1   141
#>  [36,]  44   2      1          1        1   150
#>  [37,]  43   2      0        180        1   140
#>  [38,]  45   2      0        180        1   140
#>  [39,]  46  15      0        180        0   120
#>  [40,]  46   2      1          1        0   126
#>  [41,]  47   4      1          3        0   118
#>  [42,]  45   3      0        150        0   130
#>  [43,]  44   3      1          0        1   180
#>  [44,]  46   7      1          2        0   166
#>  [45,]  43  29      0        180        1   180
#>  [46,]  45   4      1          0        0   124
#>  [47,]  47   6      1          0        1   116
#>  [48,]  46  13      1         10        0   100
#>  [49,]  49   5      0         73        1   136
#>  [50,]  45   5      0          5        0   141
#>  [51,]  46   2      1          1        1   122
#>  [52,]  44   4      1          0        1   114
#>  [53,]  47   2      0        180        0   108
#>  [54,]  44   9      1          8        1   135
#>  [55,]  45   5      0        180        1   190
#>  [56,]  44   2      0        180        0   142
#>  [57,]  45   9      1          0        1   145
#>  [58,]  48  12      1         11        0   200
#>  [59,]  47   5      1          3        1   130
#>  [60,]  46   3      1          0        1   119
#>  [61,]  49   4      0        180        0   117
#>  [62,]  47   7      0        180        0   145
#>  [63,]  49   7      1          7        1   110
#>  [64,]  46   3      1          1        1   140
#>  [65,]  46   9      1          9        1   122
#>  [66,]  50   7      0        180        1   110
#>  [67,]  51   1      0          1        1   145
#>  [68,]  47   2      0        180        0   150
#>  [69,]  49  23      0        179        1   112
#>  [70,]  46   6      1          0        1   156
#>  [71,]  52   2      0        180        1   170
#>  [72,]  50   4      0          4        1   100
#>  [73,]  51   3      1          2        0   113
#>  [74,]  50   9      0        180        0   130
#>  [75,]  49   7      1          4        1    90
#>  [76,]  47   8      0        180        0   160
#>  [77,]  47   6      0        180        1   162
#>  [78,]  52   2      0        180        0   155
#>  [79,]  46   3      0        180        1   120
#>  [80,]  46   1      1          1        0   145
#>  [81,]  48   7      1          0        1   110
#>  [82,]  47   2      1          1        0   110
#>  [83,]  49   9      1          3        0   102
#>  [84,]  49  15      0        180        1   160
#>  [85,]  53   5      0        180        1   140
#>  [86,]  53   5      0         77        0   159
#>  [87,]  54   6      1          3        0   129
#>  [88,]  50   2      0          5        1   106
#>  [89,]  50  10      1          6        0   122
#>  [90,]  50  14      1         13        0   170
#>  [91,]  53   8      1          7        0   160
#>  [92,]  48   3      1          2        0   150
#>  [93,]  51  25      1          1        0   202
#>  [94,]  53   4      0          4        0   140
#>  [95,]  52  14      1          7        1   200
#>  [96,]  48   6      0        180        0   160
#>  [97,]  48  11      1         10        0   120
#>  [98,]  51  13      0         99        1   160
#>  [99,]  54   9      1          0        1   138
#> [100,]  55   6      1          2        1   114
#> [101,]  55   4      1          2        0   150
#> [102,]  51  13      1         11        0   145
#> [103,]  50   5      1          4        1   150
#> [104,]  54   4      1          0        1   121
#> [105,]  52   4      0        180        0   183
#> [106,]  50   3      0        174        1   153
#> [107,]  55  28      1         13        1   160
#> [108,]  49   6      1          0        1   130
#> [109,]  49   1      0          1        1   110
#> [110,]  53   9      0          9        1    95
#> [111,]  50   7      1          0        1   127
#> [112,]  52   5      0        175        1   117
#> [113,]  55   1      0        180        0   127
#> [114,]  54   1      0        180        0   162
#> [115,]  54   7      1          0        1   100
#> [116,]  56   3      0        180        1   193
#> [117,]  56   2      0        180        0   132
#> [118,]  55   5      1          4        1   120
#> [119,]  52   8      0        180        0   119
#> [120,]  53  18      1          9        1   150
#> [121,]  54   3      0        180        1   180
#> [122,]  52  16      0         16        0   152
#> [123,]  53  10      1          9        0   172
#> [124,]  52  16      1         14        0   170
#> [125,]  53  15      0         15        1    90
#> [126,]  53   4      0        180        1   150
#> [127,]  55   2      0        134        1   140
#> [128,]  54   3      0        180        0   128
#> [129,]  55   1      0          2        0   130
#> [130,]  57   3      0          3        0   120
#> [131,]  54   7      1          2        0   129
#> [132,]  52   9      1          3        0   170
#> [133,]  57   1      0        180        1   156
#> [134,]  52   2      0        180        0   140
#> [135,]  55  11      1          7        0   104
#> [136,]  52  15      1         14        0   130
#> [137,]  56  14      1         11        0   130
#> [138,]  58   8      0          8        1   130
#> [139,]  54   5      0        180        1   108
#> [140,]  55   3      1          1        1   156
#> [141,]  57   0      0          0        1   150
#> [142,]  53  21      1         13        1   130
#> [143,]  59   3      1          1        0   172
#> [144,]  57   4      0        180        1   119
#> [145,]  58   6      1          0        1    90
#> [146,]  53  15      1         10        1   130
#> [147,]  54  17      1          8        1   227
#> [148,]  55   9      1          2        1   147
#> [149,]  55  13      0        166        1   140
#> [150,]  54  23      1          8        0   120
#> [151,]  53   7      1          0        1   120
#> [152,]  55   5      0          5        1   131
#> [153,]  54   7      1          0        1   141
#> [154,]  56   4      0          4        0   164
#> [155,]  59  15      1         10        0   140
#> [156,]  58   9      1          0        1   180
#> [157,]  58   1      1          1        1   200
#> [158,]  55   5      1          0        0   140
#> [159,]  56   7      1          5        1   120
#> [160,]  55   2      0          2        0   106
#> [161,]  59   9      1          1        1   125
#> [162,]  57   1      0        180        0   148
#> [163,]  59   3      0        180        0   120
#> [164,]  58   4      1          0        1   160
#> [165,]  60   5      1          1        0   138
#> [166,]  57   5      0        180        1   130
#> [167,]  58  11      1          9        1   124
#> [168,]  55   5      1          0        1   160
#> [169,]  57  10      1          9        0   103
#> [170,]  59   5      0        180        1   155
#> [171,]  59   4      1          0        1   152
#> [172,]  58  26      1          0        1   189
#> [173,]  61   9      0          9        1   160
#> [174,]  60   0      1          0        1    80
#> [175,]  58   8      0        161        1   140
#> [176,]  58  14      1          6        0   190
#> [177,]  61   4      1          3        0   151
#> [178,]  61   9      1          8        0   150
#> [179,]  61   3      1          2        1   102
#> [180,]  61  20      1         13        0   130
#> [181,]  57   2      1          0        1   116
#> [182,]  58  10      0         10        1   150
#> [183,]  61   3      0         17        0   143
#> [184,]  57   3      1          2        0   120
#> [185,]  58  19      1         13        1   140
#> [186,]  59   9      1          0        1    80
#> [187,]  55   4      1          3        1   160
#> [188,]  58  11      0        172        1   135
#> [189,]  60  12      1          0        1   114
#> [190,]  61   4      1          0        1   115
#> [191,]  56   8      1          8        0   120
#> [192,]  59  11      1          8        1   190
#> [193,]  57   1      0          1        0   126
#> [194,]  59  10      0        180        0   160
#> [195,]  61   8      0         77        0   120
#> [196,]  58   8      1          5        0   152
#> [197,]  62  10      1          0        1   153
#> [198,]  57   3      1          0        0   100
#> [199,]  61  18      0        170        0   140
#> [200,]  61  28      1          7        0   133
#> [201,]  58   8      1          3        1   150
#> [202,]  61   6      0          6        0   134
#> [203,]  57  12      1          9        1   120
#> [204,]  62   4      1          0        0   160
#> [205,]  58   3      1          0        1   146
#> [206,]  59  16      1          9        1   133
#> [207,]  63   6      0         28        1   120
#> [208,]  61  13      0         13        0   120
#> [209,]  61   5      0          5        1   110
#> [210,]  57  18      1          9        1    93
#> [211,]  61   5      0          5        1   160
#> [212,]  58  11      1          9        0   179
#> [213,]  57   2      1          1        0   159
#> [214,]  62   1      1          0        1   172
#> [215,]  58   7      0        180        1   150
#> [216,]  63   3      1          1        0   180
#> [217,]  63   1      0        180        1   130
#> [218,]  61   7      0        180        0   135
#> [219,]  63   4      1          3        0   222
#> [220,]  62   3      0        180        1   105
#> [221,]  63  15      1         10        1   126
#> [222,]  64   4      0        180        0   130
#> [223,]  59   8      0        180        1   140
#> [224,]  61   9      1          9        1   150
#> [225,]  62   7      0          7        0   150
#> [226,]  59   1      0         22        1   162
#> [227,]  58   2      0        180        0   127
#> [228,]  59   4      0        180        0   196
#> [229,]  60   7      1          5        1   141
#> [230,]  65  13      0        180        1   100
#> [231,]  63   1      0          1        0   130
#> [232,]  61  15      1         13        0   170
#> [233,]  59   4      0          4        0   149
#> [234,]  60   3      0          3        0   168
#> [235,]  59  10      0        180        1   130
#> [236,]  60   8      0         17        1   130
#> [237,]  61   6      1          1        1   117
#> [238,]  64  12      1         11        0   160
#> [239,]  66   1      1          0        1   120
#> [240,]  64   6      1          0        1   140
#> [241,]  65  36      1         11        0   140
#> [242,]  66   3      1          1        0   127
#> [243,]  61  10      1          2        1   194
#> [244,]  64  32      1          9        1   160
#> [245,]  63  12      1          9        0   114
#> [246,]  66   5      1          0        1   110
#> [247,]  65   8      1          0        0   168
#> [248,]  64   0      0          0        1    90
#> [249,]  61  12      1         11        0   154
#> [250,]  64   9      0        180        0   150
#> [251,]  61   4      0        180        1   113
#> [252,]  65   3      0        180        1   190
#> [253,]  63  16      1          7        1   110
#> [254,]  66   6      1          1        1   130
#> [255,]  63  12      0         12        1   150
#> [256,]  62   3      1          1        1   199
#> [257,]  65   6      0          9        0   112
#> [258,]  65   3      1          0        1    80
#> [259,]  63   5      1          4        0   170
#> [260,]  63   2      1          1        0   180
#> [261,]  62  13      1         11        0   180
#> [262,]  64   2      0          2        0   201
#> [263,]  66  16      1         11        1   169
#> [264,]  61  14      1          5        0   140
#> [265,]  61  15      1         10        0   130
#> [266,]  63   3      1          2        0   120
#> [267,]  63   2      1          0        0   140
#> [268,]  65   8      1          0        1   140
#> [269,]  67   6      0        180        1   170
#> [270,]  68   5      1          4        1   150
#> [271,]  64  13      1         12        1   150
#> [272,]  66  13      1          0        0   118
#> [273,]  64  14      1         13        1   150
#> [274,]  65   3      0          3        0   105
#> [275,]  64   0      0          0        1   148
#> [276,]  67   4      1          3        0   130
#> [277,]  66   6      1          0        1   140
#> [278,]  64  10      1          9        1   110
#> [279,]  67   8      1          1        1   130
#> [280,]  68   5      0          5        1    90
#> [281,]  63  10      0         16        1   160
#> [282,]  64   1      0          1        1   120
#> [283,]  63   8      1          1        1   162
#> [284,]  65  18      1          3        0   120
#> [285,]  67  11      0         11        0   150
#> [286,]  68  11      0        180        0   160
#> [287,]  65  15      1         12        1   150
#> [288,]  66  11      1          0        0   100
#> [289,]  65   4      1          2        1   145
#> [290,]  69  12      0         15        1   140
#> [291,]  66  15      1         13        1   160
#> [292,]  65  11      1          6        0   130
#> [293,]  66   9      1          8        0   130
#> [294,]  63   8      0        180        1   120
#> [295,]  68  14      1         13        1   140
#> [296,]  65   8      1          0        1    90
#> [297,]  67   1      0        180        1   160
#> [298,]  65   1      1          0        0   133
#> [299,]  67   7      1          4        1   130
#> [300,]  63   2      1          0        0    99
#> [301,]  67   2      0        180        0   184
#> [302,]  65   6      0          6        0    80
#> [303,]  65  10      1          1        1   148
#> [304,]  66  19      1         12        1   150
#> [305,]  67  12      1         12        0   160
#> [306,]  64   4      0        179        0   160
#> [307,]  64  11      0         11        0   125
#> [308,]  64   4      0        180        1   140
#> [309,]  64   0      1          0        1   118
#> [310,]  67   2      0         18        0   131
#> [311,]  66   7      1          5        1   131
#> [312,]  69   4      1          3        1   150
#> [313,]  65  13      1         12        1   130
#> [314,]  69  17      1         10        0   140
#> [315,]  69   8      0         93        0   140
#> [316,]  64  21      0         21        1   155
#> [317,]  66   6      0        180        0   140
#> [318,]  65   1      0          1        1   120
#> [319,]  68  18      1          0        1   160
#> [320,]  65   6      0        101        1   115
#> [321,]  71   3      0          5        0   112
#> [322,]  68   7      0        150        0   210
#> [323,]  66   9      1          3        1   151
#> [324,]  66   1      1          1        1   165
#> [325,]  70   4      1          0        1   180
#> [326,]  69   8      0        180        1   153
#> [327,]  70  14      0        171        0   166
#> [328,]  67   6      1          4        0   130
#> [329,]  65   2      0        180        0   130
#> [330,]  69   3      1          2        0   151
#> [331,]  67  14      1         13        0   130
#> [332,]  69   8      0        180        1   180
#> [333,]  71   7      0          7        0   230
#> [334,]  66   2      0          2        1   228
#> [335,]  71   6      0         45        1   158
#> [336,]  69   5      0          5        1   142
#> [337,]  69   3      0          3        1   130
#> [338,]  67   1      0         36        1   104
#> [339,]  67   5      0          5        0   130
#> [340,]  69   8      1          5        1   195
#> [341,]  72   3      1          0        1   132
#> [342,]  69   8      1          7        1   108
#> [343,]  67   3      0        180        0   110
#> [344,]  67  14      0        172        1   140
#> [345,]  69   4      1          3        0   132
#> [346,]  70   3      0        123        0   130
#> [347,]  70   9      0        180        1   142
#> [348,]  68   3      0         19        0   135
#> [349,]  67  12      1          8        0   120
#> [350,]  69   1      0          1        1   110
#> [351,]  67   8      1          0        1   130
#> [352,]  68  10      1          8        1   160
#> [353,]  66  24      1         13        0   130
#> [354,]  72  30      1          0        1   145
#> [355,]  70   7      0          7        0   102
#> [356,]  68   7      1          2        0   135
#> [357,]  73  20      1          0        1   170
#> [358,]  71   6      0          9        0   120
#> [359,]  70  11      0        180        1   210
#> [360,]  72  12      1         10        0   170
#> [361,]  67   8      0        180        1   170
#> [362,]  67   9      0        180        0   158
#> [363,]  73  13      0        152        1   130
#> [364,]  70   5      0        180        0   150
#> [365,]  71   1      0        173        1   188
#> [366,]  68  23      0        180        1   220
#> [367,]  70   3      0        180        0   121
#> [368,]  71   3      1          2        0   150
#> [369,]  72   5      0         28        0   120
#> [370,]  71   5      0        180        0   191
#> [371,]  73   6      0        180        1   117
#> [372,]  69  16      1         10        1   140
#> [373,]  68   7      0        180        1   130
#> [374,]  72  16      1          1        1   130
#> [375,]  70   4      0        180        0   180
#> [376,]  73   6      1          0        1   270
#> [377,]  72   8      1          1        1   150
#> [378,]  71   2      1          0        1   180
#> [379,]  73   7      0          7        1   140
#> [380,]  68  15      1         13        1   130
#> [381,]  70  13      1          9        0   100
#> [382,]  72   6      0        180        1   130
#> [383,]  73   0      0        180        1   161
#> [384,]  74   8      1          0        1    85
#> [385,]  73   4      0        180        1   154
#> [386,]  71  15      1         11        0   165
#> [387,]  74  20      0         20        1   180
#> [388,]  68   9      0        180        1   120
#> [389,]  71  20      1         10        0   140
#> [390,]  74   0      1          0        1    90
#> [391,]  70   5      1          0        1   190
#> [392,]  71   8      1          7        0   149
#> [393,]  71   3      1          2        1   190
#> [394,]  73  10      1          8        0   106
#> [395,]  69  12      1          1        1   149
#> [396,]  70  26      1         11        1   120
#> [397,]  74   4      0          4        0   120
#> [398,]  73   4      0         58        1   160
#> [399,]  73   6      0        180        0   110
#> [400,]  72   8      1          0        1   140
#> [401,]  71  13      1          8        0   121
#> [402,]  71  14      1         13        1   170
#> [403,]  72  10      1          8        1   153
#> [404,]  69   7      0        180        1   144
#> [405,]  72  15      1         13        0   156
#> [406,]  71  10      1          9        1   120
#> [407,]  75   2      1          1        0   145
#> [408,]  73  10      1          9        1   146
#> [409,]  72  10      1          9        1   160
#> [410,]  73  10      1         10        1   120
#> [411,]  74  15      1          9        1   179
#> [412,]  71   2      0         10        1   112
#> [413,]  75   9      1          7        0   140
#> [414,]  75  13      1          1        1   130
#> [415,]  72  15      1         12        1   120
#> [416,]  73  10      1          8        0   120
#> [417,]  70   7      1          4        0   184
#> [418,]  72   7      0         57        1   145
#> [419,]  72  11      0         11        1   140
#> [420,]  73   5      1          3        1   112
#> [421,]  76  25      1         12        1   170
#> [422,]  73  12      1         12        1   140
#> [423,]  72   2      0        180        0   120
#> [424,]  75   1      0        180        1   140
#> [425,]  71   3      1          0        0   144
#> [426,]  76   3      1          0        1   120
#> [427,]  71  32      1         12        1   107
#> [428,]  72   5      0        180        0   154
#> [429,]  76   5      0          5        1   130
#> [430,]  77  11      0         11        1   150
#> [431,]  77   4      0          4        0   185
#> [432,]  75   3      1          1        0   180
#> [433,]  72   7      1          2        0   142
#> [434,]  76   1      0        180        0   114
#> [435,]  74   2      1          1        0   140
#> [436,]  76   8      1          0        1   141
#> [437,]  74  19      1          4        1   200
#> [438,]  73   6      0          6        1   114
#> [439,]  75  23      1         14        1   110
#> [440,]  72   4      0         85        1   120
#> [441,]  76  17      1          0        1   200
#> [442,]  76  13      1         10        0   110
#> [443,]  75   4      1          0        1   122
#> [444,]  75   7      0          7        0   190
#> [445,]  75   0      0          0        1   130
#> [446,]  73  13      1         11        0   195
#> [447,]  75  12      0         12        1   160
#> [448,]  74   8      1          0        1   105
#> [449,]  76  13      1          8        1   148
#> [450,]  76   4      0          4        1   155
#> [451,]  75   1      0          1        1   125
#> [452,]  74   2      0        180        0   111
#> [453,]  73   1      0         52        1   105
#> [454,]  72   5      0        180        0   120
#> [455,]  76  44      1         10        0   105
#> [456,]  76   5      0        180        0   185
#> [457,]  74  10      1          0        1   135
#> [458,]  76   5      1          0        1   167
#> [459,]  74   8      1          8        1   170
#> [460,]  73  33      1         12        1   175
#> [461,]  73  10      1          9        0   146
#> [462,]  77  12      0        180        0   130
#> [463,]  77   1      1          0        1    90
#> [464,]  76  12      1         11        1   120
#> [465,]  78   5      1          0        1   170
#> [466,]  73   7      1          0        0   174
#> [467,]  74   9      1          8        0   200
#> [468,]  79  10      1          8        0   190
#> [469,]  74   2      1          0        1   130
#> [470,]  78  18      0         18        1   144
#> [471,]  77   3      0        180        0   110
#> [472,]  76  29      0         47        0    90
#> [473,]  73   8      1          1        1   162
#> [474,]  74   2      0        180        0   100
#> [475,]  78   7      0          7        1   133
#> [476,]  74  15      0        180        1   172
#> [477,]  78   8      1          6        1   110
#> [478,]  74   7      0          7        0   161
#> [479,]  78  32      1          9        1   198
#> [480,]  79   6      0        180        0   170
#> [481,]  80  10      1          6        1   147
#> [482,]  78   0      0        180        1   212
#> [483,]  75   5      0        119        1   150
#> [484,]  75  12      1          1        1   120
#> [485,]  80   8      0          8        1   120
#> [486,]  75  13      1          6        0   150
#> [487,]  74  10      1          8        0   135
#> [488,]  79   4      0         80        0   145
#> [489,]  75   4      1          0        0   212
#> [490,]  78  10      0        180        1   130
#> [491,]  76  11      1          0        0   120
#> [492,]  75   3      0          3        0     0
#> [493,]  76   7      0         29        1   150
#> [494,]  77  24      0         24        1   160
#> [495,]  79   8      0         32        1   120
#> [496,]  80   9      0         23        1   128
#> [497,]  80   6      0          6        1   150
#> [498,]  78   6      1          0        1   240
#> [499,]  76   3      1          0        1   140
#> [500,]  78  11      1          1        1   140
#> [501,]  79  11      0        180        0   160
#> [502,]  79   2      1          0        1   121
#> [503,]  78  14      1          0        1   140
#> [504,]  81   1      0          1        0   130
#> [505,]  78  11      1          8        1   118
#> [506,]  79   4      0          4        1   125
#> [507,]  76  10      1          8        0   180
#> [508,]  77   6      0          6        1   107
#> [509,]  75   2      1          1        1   204
#> [510,]  76   1      0          1        1   140
#> [511,]  78   7      1          0        1   110
#> [512,]  79   3      0          3        0   120
#> [513,]  81   1      0        180        0   120
#> [514,]  80  15      1         12        1   150
#> [515,]  77   9      1          4        0   141
#> [516,]  80  40      1          0        1   138
#> [517,]  78   4      0         59        1   112
#> [518,]  79  10      0         10        1   120
#> [519,]  80  15      1          0        1    90
#> [520,]  81   4      1          2        1   126
#> [521,]  80   6      0        173        1   160
#> [522,]  78  32      0        180        1   130
#> [523,]  79   1      0         37        1   140
#> [524,]  81   3      0        180        0   184
#> [525,]  81   2      0        175        0   172
#> [526,]  78   7      0          7        1   147
#> [527,]  77  13      1          0        1   190
#> [528,]  78  15      0         15        0   165
#> [529,]  80   5      1          1        1   108
#> [530,]  78   4      0        180        0   175
#> [531,]  81   4      1          1        1   104
#> [532,]  78  20      1          0        1   109
#> [533,]  78   3      1          1        1   152
#> [534,]  77  10      1          8        1   130
#> [535,]  82   3      1          1        1   144
#> [536,]  77   5      0         85        0   188
#> [537,]  80   6      1          0        1   119
#> [538,]  78   2      0        180        0   148
#> [539,]  80   5      0          5        1   130
#> [540,]  82   1      1          0        1    82
#> [541,]  79  10      0        180        1   150
#> [542,]  81   1      0        108        0   129
#> [543,]  79   1      0        125        0   193
#> [544,]  82  21      1          2        0   155
#> [545,]  84  22      1         10        0   180
#> [546,]  79   4      0          4        1   121
#> [547,]  82   5      0        180        0   110
#> [548,]  83   5      0        180        0   148
#> [549,]  79   7      1          6        0   130
#> [550,]  83   4      0        103        0    97
#> [551,]  81  11      1          8        0   160
#> [552,]  80  11      1          8        0   170
#> [553,]  78  23      1         10        1   145
#> [554,]  79   4      0          4        1   183
#> [555,]  78   9      1          4        1   120
#> [556,]  82   8      1          1        0   128
#> [557,]  80   7      1          0        1   146
#> [558,]  81  20      1          9        0   170
#> [559,]  81  16      0         16        1   110
#> [560,]  80  11      1          8        0   110
#> [561,]  79   0      1          0        1    96
#> [562,]  85   4      0        180        0    90
#> [563,]  81   2      1          1        0   198
#> [564,]  83   2      0          2        1   155
#> [565,]  82   6      0        128        1   100
#> [566,]  84   4      0        167        0   198
#> [567,]  84   5      0        180        1   203
#> [568,]  84   4      0          4        1    85
#> [569,]  81   1      0          1        1   150
#> [570,]  83   3      0        180        0   174
#> [571,]  81   4      0         90        1   138
#> [572,]  79   9      1          8        0   150
#> [573,]  85   3      1          2        1   160
#> [574,]  80  13      1          8        1   140
#> [575,]  80   2      1          0        1   130
#> [576,]  80   6      0         71        1   189
#> [577,]  83   1      0          1        1   100
#> [578,]  82  19      0         19        0   120
#> [579,]  80  30      1         13        0   220
#> [580,]  83   9      0        180        0   198
#> [581,]  79  14      1          0        0   110
#> [582,]  83   3      0        114        0    98
#> [583,]  81  14      1         12        1   128
#> [584,]  83   2      0        154        0   130
#> [585,]  82   0      0          2        1   100
#> [586,]  83   1      0        180        0   160
#> [587,]  82  16      1          8        0   103
#> [588,]  81   4      0          4        0   160
#> [589,]  86  12      0        180        1   120
#> [590,]  83  12      1          2        1   170
#> [591,]  81  19      1         14        0   120
#> [592,]  82   3      1          2        0   130
#> [593,]  83   7      0        126        0   135
#> [594,]  86   8      0          8        1   132
#> [595,]  81  16      1          9        0   180
#> [596,]  84   6      0        165        0   145
#> [597,]  86   3      0          3        1   140
#> [598,]  82   9      0        180        1   134
#> [599,]  84   3      0        180        1   120
#> [600,]  81  13      0        180        0   152
#> [601,]  85   3      0          3        1   118
#> [602,]  81   4      0        180        0   160
#> [603,]  83   9      0        180        1   149
#> [604,]  82   1      0        180        1   193
#> [605,]  87   2      0          5        1   137
#> [606,]  86  12      1          0        1   132
#> [607,]  86   6      1          0        1   140
#> [608,]  84  16      0         70        1   150
#> [609,]  84   3      1          2        0   125
#> [610,]  84   3      0          3        1   121
#> [611,]  83  13      1         12        0   170
#> [612,]  84   7      1          2        0   148
#> [613,]  87   2      0        180        0   113
#> [614,]  84   9      0         92        1   110
#> [615,]  84   3      0        180        1   170
#> [616,]  86   6      0          6        1   117
#> [617,]  84  13      0         62        1   100
#> [618,]  86   6      1          1        0   112
#> [619,]  88   4      0          4        0   100
#> [620,]  83  20      1          3        1   150
#> [621,]  88   4      0          4        1   115
#> [622,]  85  22      0         22        1   184
#> [623,]  88   3      0        115        0   110
#> [624,]  88   2      0        180        1    68
#> [625,]  83   3      0          3        1   130
#> [626,]  86  15      1          8        1   109
#> [627,]  88   4      0          4        0    86
#> [628,]  89   4      0          4        1   153
#> [629,]  87   6      0        180        1   110
#> [630,]  84   8      0        180        1   119
#> [631,]  85   8      0          8        1   136
#> [632,]  87  15      1          9        1   138
#> [633,]  89  10      0         46        1   170
#> [634,]  88   1      0          1        0   135
#> [635,]  86   4      0        180        1   145
#> [636,]  87   6      1          0        0   125
#> [637,]  91  10      0        145        0   135
#> [638,]  88   8      0         50        1   154
#> [639,]  90  11      1         10        1   186
#> [640,]  87   6      0        126        1   168
#> [641,]  86  10      0        180        1   137
#> [642,]  86   9      1          7        0   130
#> [643,]  90   4      1          0        0   121
#> [644,]  87  43      0        178        1   130
#> [645,]  87   5      0         36        1   150
#> [646,]  88   3      1          2        0   159
#> [647,]  89   3      1          1        1   160
#> [648,]  87   7      0         74        1   105
#> [649,]  89  12      1          0        1   130
#> [650,]  89   2      0        168        0   118
#> [651,]  91   5      0        169        1   176
#> [652,]  92   7      0          7        1   110
#> [653,]  89   4      0          4        1   159
#> [654,]  91   0      0          0        0     0
#> [655,]  89  14      0        180        1    84
#> [656,]  90  18      0        180        0   188
#> [657,]  91   4      1          0        1   120
#> [658,]  94   6      0         50        0    78
#> [659,]  91   2      0          2        1   116
#> [660,]  94   8      0          8        1   142
#> [661,]  92   4      0         76        1   149
#> [662,]  96   3      0         12        1    97
#> [663,]  95   8      1          5        1   150
#> [664,]  94   3      0         26        1   144
#> [665,]  91  12      0         53        1   212
#> [666,]  93   0      1          0        1   122
#> [667,]  92   5      0         69        0   139
#> [668,]  92   2      0          2        0   112
#> [669,]  96   3      1          0        1   104
#> [670,]  96  15      1          0        1   140
#> 
#> $y
#>   [1] 180.0+   5.0+ 180.0+   5.0+   2.0+ 180.0+ 180.0+ 115.0  180.0+  12.0 
#>  [11]   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 180.0+ 180.0+   3.0 
#>  [21] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 180.0+ 155.0+
#>  [31] 180.0+ 180.0+ 180.0+ 180.0+   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [41] 180.0+ 150.0  180.0+ 180.0+ 180.0+ 180.0+   6.0+ 180.0+  73.0    5.0+
#>  [51] 161.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 177.0+ 180.0+ 180.0+ 180.0+
#>  [61] 180.0+ 180.0+   7.0  180.0+ 180.0+ 180.0+   1.0  180.0+ 179.0+ 180.0+
#>  [71] 180.0+   4.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [81]   7.0  180.0+ 180.0+ 180.0+ 180.0+  77.0  180.0+   5.0  180.0+ 180.0+
#>  [91] 180.0+ 180.0+ 180.0+   4.0+  85.0  180.0+ 180.0+  99.0  180.0+   6.0+
#> [101] 180.0+  13.0+ 171.0+ 180.0+ 180.0+ 174.0+  28.0    6.0+   1.0    9.0+
#> [111] 180.0+ 175.0+ 180.0+ 180.0+   7.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [121] 180.0+  16.0+ 180.0+  16.0   15.0+ 180.0+ 134.0+ 180.0+   2.0    3.0+
#> [131] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   8.0+ 180.0+ 180.0+
#> [141]   0.5  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 171.0+  15.0  166.0+ 180.0+
#> [151] 180.0+   5.0+ 180.0+   4.0+ 180.0+   9.0+   1.0  180.0+ 180.0+   2.0+
#> [161] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [171] 180.0+ 180.0+   9.0+   0.5  161.0+ 171.0+ 180.0+ 180.0+   3.0  180.0+
#> [181] 180.0+  10.0+  17.0    3.0+  19.0    9.0+ 180.0+ 172.0+ 172.0+ 180.0+
#> [191]   8.0  180.0+   1.0+ 180.0+  77.0    8.0+ 180.0+ 180.0+ 170.0   94.0 
#> [201] 180.0+   6.0  180.0+ 180.0+   3.0+ 180.0+  28.0   13.0+   5.0   18.0 
#> [211]   5.0+ 180.0+ 180.0+   1.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [221] 180.0+ 180.0+ 180.0+ 180.0+   7.0+  22.0  180.0+ 180.0+  84.0  180.0+
#> [231]   1.0  180.0+   4.0+   3.0+ 180.0+  17.0  180.0+  12.0  180.0+ 180.0+
#> [241]  36.0    3.0+  88.0  180.0+  12.0  180.0+ 180.0+   0.5   12.0+ 180.0+
#> [251] 180.0+ 180.0+ 180.0+ 180.0+  12.0  180.0+   9.0    3.0  180.0+ 180.0+
#> [261] 180.0+   2.0+ 180.0+ 180.0+ 180.0+   3.0+   2.0+  15.0  180.0+   5.0+
#> [271]  13.0  166.0+  14.0+   3.0    0.5+ 180.0+ 180.0+ 180.0+   8.0    5.0 
#> [281]  16.0    1.0  180.0+ 123.0+  11.0+ 180.0+  15.0+ 180.0+   4.0+  15.0 
#> [291] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   8.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [301] 180.0+   6.0  180.0+  19.0+  12.0  179.0+  11.0+ 180.0+   0.5   18.0 
#> [311]   7.0+ 152.0+ 180.0+ 180.0+  93.0   21.0+ 180.0+   1.0   18.0+ 101.0 
#> [321]   5.0  150.0  180.0+   1.0  180.0+ 180.0+ 171.0    6.0  180.0+ 180.0+
#> [331] 180.0+ 180.0+   7.0+   2.0   45.0    5.0+   3.0+  36.0    5.0+ 180.0+
#> [341] 180.0+   8.0+ 180.0+ 172.0+ 180.0+ 123.0  180.0+  19.0  180.0+   1.0 
#> [351] 180.0+  10.0+ 180.0+ 162.0    7.0+   7.0+ 124.0    9.0  180.0+  12.0 
#> [361] 180.0+ 180.0+ 152.0  180.0+ 173.0+ 180.0+ 180.0+ 180.0+  28.0  180.0+
#> [371] 180.0+  16.0+ 180.0+  16.0+ 180.0+   6.0  180.0+ 180.0+   7.0+  15.0 
#> [381]  13.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  20.0  180.0+  20.0    0.5 
#> [391] 180.0+   8.0    3.0   87.0   12.0  180.0+   4.0+  58.0  180.0+ 180.0+
#> [401] 175.0   14.0+  10.0+ 180.0+ 180.0+ 179.0+ 180.0+ 180.0+ 159.0   15.0 
#> [411] 180.0+  10.0  180.0+  13.0  180.0+  10.0  104.0+  57.0   11.0    5.0 
#> [421] 180.0+  12.0  180.0+ 180.0+ 180.0+ 180.0+ 177.0+ 180.0+   5.0   11.0+
#> [431]   4.0+ 180.0+   7.0  180.0+ 180.0+ 180.0+ 180.0+   6.0  180.0+  85.0 
#> [441]  17.0+ 174.0+   4.0    7.0    0.5  180.0+  12.0  180.0+ 180.0+   4.0 
#> [451]   1.0  180.0+  52.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+   8.0   33.0 
#> [461] 180.0+ 180.0+   1.0   12.0  180.0+   7.0+ 168.0+ 180.0+ 176.0+  18.0 
#> [471] 180.0+  47.0  180.0+ 180.0+   7.0  180.0+   8.0+   7.0   32.0  180.0+
#> [481]  10.0  180.0+ 119.0   12.0    8.0  180.0+ 180.0+  80.0    4.0+ 180.0+
#> [491]  11.0    3.0   29.0   24.0   32.0   23.0    6.0  180.0+   3.0+ 180.0+
#> [501] 180.0+ 180.0+ 180.0+   1.0   11.0    4.0   10.0+   6.0    2.0+   1.0 
#> [511]  43.0    3.0  180.0+ 180.0+  71.0   40.0   59.0   10.0+ 180.0+  93.0 
#> [521] 173.0  180.0+  37.0  180.0+ 175.0+   7.0+  22.0   15.0+   5.0+ 180.0+
#> [531]  71.0   20.0+   3.0+  10.0  180.0+  85.0    6.0  180.0+   5.0    1.0 
#> [541] 180.0+ 108.0  125.0  180.0+ 180.0+   4.0  180.0+ 180.0+ 180.0+ 103.0 
#> [551] 180.0+ 169.0   70.0    4.0  180.0+ 180.0+   7.0+  20.0   16.0  180.0+
#> [561]   0.5  180.0+ 180.0+   2.0  128.0  167.0  180.0+   4.0    1.0  180.0+
#> [571]  90.0  180.0+ 180.0+ 180.0+ 180.0+  71.0    1.0   19.0   30.0  180.0+
#> [581] 180.0+ 114.0  180.0+ 154.0    2.0  180.0+  16.0+   4.0+ 180.0+  77.0 
#> [591] 180.0+   3.0  126.0    8.0  180.0+ 165.0    3.0  180.0+ 180.0+ 180.0+
#> [601]   3.0+ 180.0+ 180.0+ 180.0+   5.0  180.0+   6.0   70.0  180.0+   3.0 
#> [611]  13.0  180.0+ 180.0+  92.0  180.0+   6.0+  62.0    6.0+   4.0+  20.0 
#> [621]   4.0   22.0  115.0  180.0+   3.0+ 180.0+   4.0    4.0  180.0+ 180.0+
#> [631]   8.0  180.0+  46.0    1.0+ 180.0+  25.0  145.0   50.0   11.0  126.0 
#> [641] 180.0+ 180.0+   4.0  178.0+  36.0   75.0    3.0+  74.0  180.0+ 168.0 
#> [651] 169.0    7.0    4.0    0.5  180.0+ 180.0+   4.0   50.0    2.0    8.0+
#> [661]  76.0   12.0    8.0   26.0   53.0    0.5   69.0    2.0    3.0   15.0+
#> 
#> $weights
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Warning: Multiple lambdas have been fit. Lambda will be set to 0.01 (see parameter 's').

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8627006