GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.coxnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.coxnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Caution: This learner is different to learners calling glmnet::cv.glmnet() in that it does not use the internal optimization of parameter lambda. Instead, lambda needs to be tuned by the user (e.g., via mlr3tuning). When lambda is tuned, the glmnet will be trained for each tuning iteration. While fitting the whole path of lambdas would be more efficient, as is done by default in glmnet::glmnet(), tuning/selecting the parameter at prediction time (using parameter s) is currently not supported in mlr3 (at least not in efficient manner). Tuning the s parameter is, therefore, currently discouraged.

When the data are i.i.d. and efficiency is key, we recommend using the respective auto-tuning counterpart in mlr_learners_surv.cv_glmnet(). However, in some situations this is not applicable, usually when data are imbalanced or not i.i.d. (longitudinal, time-series) and tuning requires custom resampling strategies (blocked design, stratification).

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.coxnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exact	logical	FALSE	TRUE, FALSE	-

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	TRUE	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	0.01		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.logistic	character	Newton	Newton, modified.Newton	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvGlmnet

Methods

Public methods

LearnerSurvGlmnet$new()

LearnerSurvGlmnet$selected_features()

LearnerSurvGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.glmnet")
print(learner)
#> 
#> ── <LearnerSurvGlmnet> (surv.glmnet): Regularized Generalized Linear Model ─────
#> • Model: -
#> • Parameters: use_pred_offset=TRUE
#> • Packages: mlr3, mlr3proba, mlr3extralearners, and glmnet
#> • Predict Types: [crank], distr, and lp
#> • Feature Types: logical, integer, and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: offset, selected_features, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#>    Df  %Dev   Lambda
#> 1   0  0.00 0.176300
#> 2   1  0.57 0.160700
#> 3   2  1.46 0.146400
#> 4   2  2.29 0.133400
#> 5   3  4.60 0.121500
#> 6   3  7.35 0.110700
#> 7   3  9.37 0.100900
#> 8   3 10.93 0.091940
#> 9   3 12.17 0.083770
#> 10  3 13.17 0.076330
#> 11  3 13.99 0.069550
#> 12  3 14.67 0.063370
#> 13  3 15.24 0.057740
#> 14  3 15.71 0.052610
#> 15  3 16.11 0.047940
#> 16  3 16.45 0.043680
#> 17  4 16.74 0.039800
#> 18  5 17.02 0.036260
#> 19  5 17.26 0.033040
#> 20  5 17.46 0.030100
#> 21  5 17.63 0.027430
#> 22  6 17.78 0.024990
#> 23  6 17.91 0.022770
#> 24  6 18.01 0.020750
#> 25  6 18.10 0.018910
#> 26  6 18.18 0.017230
#> 27  6 18.24 0.015700
#> 28  6 18.30 0.014300
#> 29  6 18.34 0.013030
#> 30  6 18.38 0.011870
#> 31  6 18.41 0.010820
#> 32  6 18.44 0.009858
#> 33  6 18.46 0.008982
#> 34  6 18.48 0.008184
#> 35  6 18.50 0.007457
#> 36  6 18.51 0.006795
#> 37  6 18.52 0.006191
#> 38  6 18.53 0.005641
#> 39  6 18.54 0.005140
#> 40  6 18.54 0.004683
#> 41  6 18.55 0.004267
#> 42  6 18.55 0.003888
#> 43  6 18.56 0.003543
#> 44  6 18.56 0.003228
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  32   5      1          0        1   121
#>   [2,]  33   2      0          2        0   150
#>   [3,]  35   2      1          1        1   112
#>   [4,]  38  13      1          0        1   161
#>   [5,]  38   2      0        115        0   150
#>   [6,]  36   1      0        180        1   155
#>   [7,]  38  12      1          8        1   120
#>   [8,]  36   5      1          0        1   115
#>   [9,]  33   6      1          1        1   115
#>  [10,]  38  16      1         10        0   160
#>  [11,]  38  12      1         11        1    92
#>  [12,]  40  12      1          9        0   153
#>  [13,]  42   3      1          1        1   130
#>  [14,]  38   5      1          3        0   125
#>  [15,]  42   2      0          2        0   140
#>  [16,]  40  11      1         10        1   120
#>  [17,]  42   2      0        180        0   100
#>  [18,]  43   3      1          0        1   100
#>  [19,]  41   2      1          1        0   166
#>  [20,]  43   4      1          0        1   130
#>  [21,]  40   3      1          1        0   170
#>  [22,]  43   2      1          1        1   116
#>  [23,]  42   2      0        180        1   124
#>  [24,]  44   5      1          1        0   170
#>  [25,]  45   3      0        180        1   154
#>  [26,]  41  10      1          8        0   150
#>  [27,]  41  13      1          1        0   140
#>  [28,]  46  15      0        180        0   120
#>  [29,]  46   2      1          1        0   126
#>  [30,]  47   4      1          3        0   118
#>  [31,]  48  15      0        180        1   160
#>  [32,]  44   3      1          0        1   180
#>  [33,]  46   7      1          2        0   166
#>  [34,]  43  10      0        180        0   185
#>  [35,]  47   6      1          0        1   116
#>  [36,]  46  13      1         10        0   100
#>  [37,]  44   0      1          0        1    96
#>  [38,]  43   3      1          0        1   124
#>  [39,]  45   8      1          0        1   117
#>  [40,]  45   5      0          5        0   141
#>  [41,]  46   2      1          1        1   122
#>  [42,]  47   2      0        180        0   108
#>  [43,]  45   5      0        180        1   190
#>  [44,]  46   5      1          3        0   130
#>  [45,]  44   2      0        180        0   142
#>  [46,]  46  15      0        180        1   120
#>  [47,]  47   3      1          1        1   120
#>  [48,]  48   3      0        180        0   154
#>  [49,]  47   5      1          3        1   130
#>  [50,]  47   9      1          6        0   170
#>  [51,]  49   4      0        180        0   117
#>  [52,]  50   1      1          0        1   129
#>  [53,]  48   2      1          0        0   184
#>  [54,]  47   7      0        180        0   145
#>  [55,]  50   6      1          2        1   140
#>  [56,]  46   3      1          1        1   140
#>  [57,]  46   9      1          9        1   122
#>  [58,]  50   7      0        180        1   110
#>  [59,]  49   2      0          2        0   105
#>  [60,]  51   1      0          1        1   145
#>  [61,]  49  15      1         11        1   160
#>  [62,]  47   2      0        180        0   150
#>  [63,]  46   6      1          0        1   156
#>  [64,]  52   2      0        180        1   170
#>  [65,]  50   7      1          0        1    92
#>  [66,]  50   4      0          4        1   100
#>  [67,]  51   3      1          2        0   113
#>  [68,]  49   7      1          4        1    90
#>  [69,]  47   6      0        180        1   162
#>  [70,]  51   8      0        180        1   140
#>  [71,]  46   3      0        180        1   120
#>  [72,]  50   4      1          1        0   150
#>  [73,]  53   8      0         36        1   160
#>  [74,]  48  17      1         10        0   111
#>  [75,]  47   2      1          1        0   110
#>  [76,]  52   4      1          4        0   152
#>  [77,]  49   9      1          3        0   102
#>  [78,]  49  15      0        180        1   160
#>  [79,]  53   5      0        180        1   140
#>  [80,]  54  17      1         12        1   102
#>  [81,]  54   6      1          3        0   129
#>  [82,]  51   3      1          1        0   140
#>  [83,]  50  10      1          6        0   122
#>  [84,]  50  14      1         13        0   170
#>  [85,]  53   8      1          7        0   160
#>  [86,]  48   3      1          2        0   150
#>  [87,]  51  25      1          1        0   202
#>  [88,]  52  14      1          7        1   200
#>  [89,]  48   6      0        180        0   160
#>  [90,]  54   9      1          0        1   138
#>  [91,]  55   3      1          1        0   150
#>  [92,]  54  23      1         10        0   131
#>  [93,]  55   6      1          2        1   114
#>  [94,]  54   9      1          1        0   130
#>  [95,]  55   4      1          2        0   150
#>  [96,]  51  13      1         11        0   145
#>  [97,]  50   5      1          4        1   150
#>  [98,]  54   4      1          0        1   121
#>  [99,]  52   4      0        180        0   183
#> [100,]  50   3      0        174        1   153
#> [101,]  50   7      1          1        0   156
#> [102,]  53   8      1          0        1   130
#> [103,]  50   7      1          0        1   127
#> [104,]  52   5      0        175        1   117
#> [105,]  55   1      0        180        0   127
#> [106,]  55   2      0          2        0   145
#> [107,]  54   1      0        180        0   162
#> [108,]  56   3      0        180        1   193
#> [109,]  55   5      1          4        1   120
#> [110,]  52   8      0        180        0   119
#> [111,]  54   3      0        180        1   180
#> [112,]  55   6      0        180        0   170
#> [113,]  52  16      0         16        0   152
#> [114,]  53  15      0         15        1    90
#> [115,]  53   4      0        180        1   150
#> [116,]  55   6      0        180        1   100
#> [117,]  55   6      1          5        1   138
#> [118,]  54  12      1          0        1   190
#> [119,]  54   3      0        180        0   128
#> [120,]  56   3      0          8        1   139
#> [121,]  55   1      0          2        0   130
#> [122,]  57   3      0          3        0   120
#> [123,]  54   7      1          2        0   129
#> [124,]  54   2      1          1        0   135
#> [125,]  57   5      1          3        1   138
#> [126,]  57   1      0        180        1   156
#> [127,]  57   1      0          1        1   100
#> [128,]  56   4      1          0        1   140
#> [129,]  55  11      1          7        0   104
#> [130,]  52  15      1         14        0   130
#> [131,]  56  14      1         11        0   130
#> [132,]  58   8      0          8        1   130
#> [133,]  54   5      0        180        1   108
#> [134,]  55   3      1          1        1   156
#> [135,]  57   0      0          0        1   150
#> [136,]  53  21      1         13        1   130
#> [137,]  59   3      1          1        0   172
#> [138,]  57   4      0        180        1   119
#> [139,]  58   6      1          0        1    90
#> [140,]  54  17      1          8        1   227
#> [141,]  55   9      1          2        1   147
#> [142,]  55  13      0        166        1   140
#> [143,]  54  23      1          8        0   120
#> [144,]  57   4      1          2        1   185
#> [145,]  53   4      0        147        1   145
#> [146,]  57  11      1         10        1   129
#> [147,]  55   5      0          5        1   131
#> [148,]  59  15      1         10        0   140
#> [149,]  58   9      1          0        1   180
#> [150,]  58   1      1          1        1   200
#> [151,]  55   5      1          0        0   140
#> [152,]  59   9      1          1        1   125
#> [153,]  57   1      0        180        0   148
#> [154,]  60  11      1          9        0   106
#> [155,]  57   2      0          2        1   120
#> [156,]  60   5      1          1        0   138
#> [157,]  57   5      0        180        1   130
#> [158,]  58  11      1          9        1   124
#> [159,]  57  10      1          9        0   103
#> [160,]  59   6      1          0        1   140
#> [161,]  59   5      0        180        1   155
#> [162,]  59   4      1          0        1   152
#> [163,]  58  26      1          0        1   189
#> [164,]  61   9      0          9        1   160
#> [165,]  58   4      1          3        0   120
#> [166,]  60   0      1          0        1    80
#> [167,]  58  14      1          6        0   190
#> [168,]  61   9      1          8        0   150
#> [169,]  58   1      0          1        1   100
#> [170,]  57  13      1         10        0   110
#> [171,]  57   2      1          0        1   116
#> [172,]  58  10      0         10        1   150
#> [173,]  57  11      0        180        1   150
#> [174,]  56  14      0         45        0   130
#> [175,]  57   3      1          2        0   120
#> [176,]  58  19      1         13        1   140
#> [177,]  56  13      1          6        1   158
#> [178,]  56  18      1         11        1   165
#> [179,]  59   9      1          0        1    80
#> [180,]  58  11      0        172        1   135
#> [181,]  60  12      1          0        1   114
#> [182,]  55   9      1          7        1   135
#> [183,]  61   4      1          0        1   115
#> [184,]  56   8      1          8        0   120
#> [185,]  61  13      1         12        1   130
#> [186,]  59  11      1          8        1   190
#> [187,]  57   1      0          1        0   126
#> [188,]  59   5      1          2        0   182
#> [189,]  58   5      1          1        1   135
#> [190,]  59  10      0        180        0   160
#> [191,]  61   8      0         77        0   120
#> [192,]  58   8      1          5        0   152
#> [193,]  62  10      1          0        1   153
#> [194,]  62   7      1          2        1   180
#> [195,]  57   3      1          0        0   100
#> [196,]  61  18      0        170        0   140
#> [197,]  61  28      1          7        0   133
#> [198,]  61   7      0          7        1   150
#> [199,]  59  13      1          2        0   198
#> [200,]  57  12      1          9        1   120
#> [201,]  62   4      1          3        0   173
#> [202,]  59   1      0        180        0   155
#> [203,]  59  16      1          9        1   133
#> [204,]  63   6      0         28        1   120
#> [205,]  61   5      0          5        1   110
#> [206,]  57  18      1          9        1    93
#> [207,]  58  11      1          9        0   179
#> [208,]  57   2      1          1        0   159
#> [209,]  62  17      1         10        1   180
#> [210,]  58   7      0        180        1   150
#> [211,]  63   3      1          1        0   180
#> [212,]  63   1      0        180        1   130
#> [213,]  63   4      0        180        1   190
#> [214,]  59   8      0        180        1   140
#> [215,]  61   9      1          9        1   150
#> [216,]  59   1      0         22        1   162
#> [217,]  58   2      0        180        0   127
#> [218,]  59   4      0        180        0   196
#> [219,]  60   7      1          5        1   141
#> [220,]  65  13      0        180        1   100
#> [221,]  63   1      0          1        0   162
#> [222,]  63   1      0          1        0   130
#> [223,]  61  15      1         13        0   170
#> [224,]  59   4      0          4        0   149
#> [225,]  60   3      0          3        0   168
#> [226,]  59  10      0        180        1   130
#> [227,]  61   6      1          1        1   117
#> [228,]  64  12      1         11        0   160
#> [229,]  66   1      1          0        1   120
#> [230,]  64   6      1          0        1   140
#> [231,]  63  10      1          0        1   148
#> [232,]  65  36      1         11        0   140
#> [233,]  63   4      1          3        0   162
#> [234,]  66   3      1          1        0   127
#> [235,]  64  32      1          9        1   160
#> [236,]  63  12      1          9        0   114
#> [237,]  63   7      0        180        0   120
#> [238,]  66   5      1          0        1   110
#> [239,]  65   8      1          0        0   168
#> [240,]  60   6      0        180        0   130
#> [241,]  61  12      1         11        0   154
#> [242,]  64   9      0        180        0   150
#> [243,]  61   4      0        180        1   113
#> [244,]  63  16      1          7        1   110
#> [245,]  64   7      0        180        1   120
#> [246,]  66   6      1          1        1   130
#> [247,]  63  12      0         12        1   150
#> [248,]  65   6      0          9        0   112
#> [249,]  63   5      1          4        0   170
#> [250,]  63   2      1          1        0   180
#> [251,]  62  13      1         11        0   180
#> [252,]  64   2      0          2        0   201
#> [253,]  66  18      1          5        0   142
#> [254,]  66  16      1         11        1   169
#> [255,]  61  14      1          5        0   140
#> [256,]  61  15      1         10        0   130
#> [257,]  63   3      1          2        0   120
#> [258,]  63   2      1          0        0   140
#> [259,]  64  19      1          8        1   160
#> [260,]  67   6      0        180        1   170
#> [261,]  65  15      1         11        1   160
#> [262,]  64  13      1         12        1   150
#> [263,]  66  13      1          0        0   118
#> [264,]  65   3      0          3        0   105
#> [265,]  64   0      0          0        1   148
#> [266,]  66   3      1          0        1   135
#> [267,]  66   6      1          0        1   140
#> [268,]  65   2      1          1        1   170
#> [269,]  68   1      0        180        1   166
#> [270,]  64  10      1          9        1   110
#> [271,]  63   7      1          0        0   162
#> [272,]  67   8      1          1        1   130
#> [273,]  68   5      0          5        1    90
#> [274,]  63  10      0         16        1   160
#> [275,]  66  14      0        180        0   130
#> [276,]  64   1      0          1        1   120
#> [277,]  68  18      0        180        1   260
#> [278,]  65  17      1         14        1   100
#> [279,]  65  18      1          3        0   120
#> [280,]  63   1      1          0        1   155
#> [281,]  68  11      0        180        0   160
#> [282,]  65  15      1         12        1   150
#> [283,]  65   4      1          2        1   145
#> [284,]  66  15      1         13        1   160
#> [285,]  65  11      1          6        0   130
#> [286,]  69  21      1         10        0   180
#> [287,]  69   6      0        180        1   100
#> [288,]  66   9      1          8        0   130
#> [289,]  63   8      0        180        1   120
#> [290,]  65   8      1          0        1    90
#> [291,]  66   3      0          3        1   138
#> [292,]  67   1      0        180        1   160
#> [293,]  68  10      1         10        1   150
#> [294,]  63   2      1          0        0    99
#> [295,]  65   6      0          6        0    80
#> [296,]  65  10      1          1        1   148
#> [297,]  66  19      1         12        1   150
#> [298,]  67  12      1         12        0   160
#> [299,]  64   4      0        179        0   160
#> [300,]  66   4      0        180        1   130
#> [301,]  70  15      1         12        1   132
#> [302,]  64  11      0         11        0   125
#> [303,]  64   0      1          0        1   118
#> [304,]  67   2      0         18        0   131
#> [305,]  66   7      1          5        1   131
#> [306,]  66   4      0        180        0   177
#> [307,]  68   4      1          0        1   160
#> [308,]  65  13      1         12        1   130
#> [309,]  69  17      1         10        0   140
#> [310,]  64  21      0         21        1   155
#> [311,]  65   1      0          1        1   120
#> [312,]  68  18      1          0        1   160
#> [313,]  68   4      0          4        1   190
#> [314,]  70   7      1          0        1   190
#> [315,]  68   7      0        150        0   210
#> [316,]  66   9      1          3        1   151
#> [317,]  70   4      1          0        1   180
#> [318,]  69   8      0        180        1   153
#> [319,]  70  14      0        171        0   166
#> [320,]  66   4      0        180        0   130
#> [321,]  67   6      1          4        0   130
#> [322,]  68  18      1         14        1   170
#> [323,]  69   0      0          0        1   148
#> [324,]  65   2      0        180        0   130
#> [325,]  68   7      1          0        1   150
#> [326,]  69   3      1          2        0   151
#> [327,]  69   8      0        180        1   180
#> [328,]  66   2      0          2        1   228
#> [329,]  71   6      0         45        1   158
#> [330,]  69   5      0          5        1   142
#> [331,]  71   3      0        103        0   133
#> [332,]  69   3      0          3        1   130
#> [333,]  70  22      1         13        0   103
#> [334,]  69   8      1          5        1   195
#> [335,]  69   6      1          4        1   174
#> [336,]  72   3      1          0        1   132
#> [337,]  69   8      1          7        1   108
#> [338,]  67   3      0        180        0   110
#> [339,]  69  19      0        180        0   130
#> [340,]  68  18      0         18        1   100
#> [341,]  67  14      0        172        1   140
#> [342,]  69  11      1          0        1   120
#> [343,]  66   2      0        180        0   130
#> [344,]  67   7      1          4        0   122
#> [345,]  69   4      1          3        0   132
#> [346,]  68   2      0          7        1   130
#> [347,]  67  13      1          9        0   130
#> [348,]  70   3      0        123        0   130
#> [349,]  70   9      0        180        1   142
#> [350,]  67  22      1          1        1   140
#> [351,]  68   3      0         19        0   135
#> [352,]  69   1      0          1        1   110
#> [353,]  67   4      0         60        1   136
#> [354,]  69   5      0         76        0   120
#> [355,]  72  13      1         11        1   195
#> [356,]  68  10      1          8        1   160
#> [357,]  72  30      1          0        1   145
#> [358,]  70   7      0          7        0   102
#> [359,]  68   7      1          2        0   135
#> [360,]  71   6      0          9        0   120
#> [361,]  72  12      1         10        0   170
#> [362,]  67   8      0        180        1   170
#> [363,]  67   5      1          0        1   147
#> [364,]  67   9      0        180        0   158
#> [365,]  71   1      0        173        1   188
#> [366,]  68  23      0        180        1   220
#> [367,]  69   3      0        180        0   220
#> [368,]  71   3      1          2        0   150
#> [369,]  68   4      1          3        0   210
#> [370,]  72   5      0         28        0   120
#> [371,]  73   6      0        180        1   117
#> [372,]  72  16      1          1        1   130
#> [373,]  69   1      1          0        0   155
#> [374,]  73   6      1          0        1   270
#> [375,]  72   8      1          1        1   150
#> [376,]  71   2      1          0        1   180
#> [377,]  68  15      1         13        1   130
#> [378,]  70   3      0          3        1   159
#> [379,]  72   6      0        180        1   130
#> [380,]  73   0      0        180        1   161
#> [381,]  73   4      0        180        1   154
#> [382,]  69   2      1          0        1   110
#> [383,]  71   3      1          1        0   150
#> [384,]  71  15      1         11        0   165
#> [385,]  74  20      0         20        1   180
#> [386,]  68   9      0        180        1   120
#> [387,]  71  20      1         10        0   140
#> [388,]  74   0      1          0        1    90
#> [389,]  73   3      1          0        1   136
#> [390,]  71  17      1         11        0   160
#> [391,]  71   8      1          7        0   149
#> [392,]  71   3      1          2        1   190
#> [393,]  69  12      1          1        1   149
#> [394,]  70  26      1         11        1   120
#> [395,]  72   5      1          3        1   160
#> [396,]  70   3      0        180        1   154
#> [397,]  73   6      0        180        0   110
#> [398,]  72  15      1          0        1   150
#> [399,]  72   8      1          0        1   140
#> [400,]  74   3      0          3        1   150
#> [401,]  71  13      1          8        0   121
#> [402,]  71  14      1         13        1   170
#> [403,]  74   7      1          0        1   117
#> [404,]  72  10      1          8        1   153
#> [405,]  69   7      0        180        1   144
#> [406,]  72  15      1         13        0   156
#> [407,]  70   8      0          8        0   120
#> [408,]  71  10      1          9        1   120
#> [409,]  75   1      0          1        0   133
#> [410,]  75   2      1          1        0   145
#> [411,]  73  10      1          9        1   146
#> [412,]  73  10      1         10        1   120
#> [413,]  74  15      1          9        1   179
#> [414,]  71   2      0         10        1   112
#> [415,]  75   9      1          7        0   140
#> [416,]  71   4      0          4        0   134
#> [417,]  72  15      1         12        1   120
#> [418,]  73  10      1          8        0   120
#> [419,]  72   7      0         57        1   145
#> [420,]  73  10      0        180        0   162
#> [421,]  72  11      0         11        1   140
#> [422,]  70   3      0          3        0   150
#> [423,]  76  25      1         12        1   170
#> [424,]  72   4      1          0        1   197
#> [425,]  71   3      1          0        0   144
#> [426,]  73   5      0        180        0   126
#> [427,]  73   4      0        180        0   124
#> [428,]  76   3      1          0        1   120
#> [429,]  71  32      1         12        1   107
#> [430,]  72   3      0        180        0   160
#> [431,]  77   4      0          4        0   185
#> [432,]  72   7      1          2        0   142
#> [433,]  73  15      0         15        1   160
#> [434,]  71  16      0        180        0   140
#> [435,]  74   7      0        180        1   150
#> [436,]  74   2      1          1        0   140
#> [437,]  76   8      1          0        1   141
#> [438,]  75  23      1         14        1   110
#> [439,]  74   2      0        180        0   190
#> [440,]  72   4      0         85        1   120
#> [441,]  72   4      1          3        0   160
#> [442,]  76  17      1          0        1   200
#> [443,]  75   4      1          0        1   122
#> [444,]  75   7      0          7        0   190
#> [445,]  75   0      0          0        1   130
#> [446,]  75  12      0         12        1   160
#> [447,]  74   8      1          0        1   105
#> [448,]  76  13      1          8        1   148
#> [449,]  75   4      1          2        1   188
#> [450,]  74   6      0        180        0   160
#> [451,]  76   4      0          4        1   155
#> [452,]  74   2      0        180        0   111
#> [453,]  73   0      0        180        0   156
#> [454,]  78  12      1         11        1   160
#> [455,]  76  44      1         10        0   105
#> [456,]  76   5      0        180        0   185
#> [457,]  74  10      1          0        1   135
#> [458,]  76   5      1          0        1   167
#> [459,]  75   9      0        180        1   140
#> [460,]  77   5      1          0        0   123
#> [461,]  73  10      1          9        0   146
#> [462,]  77  12      0        180        0   130
#> [463,]  76  12      1         11        1   120
#> [464,]  78   5      1          0        1   170
#> [465,]  73   7      1          0        0   174
#> [466,]  74   6      0         79        1   140
#> [467,]  75   3      1          1        1   171
#> [468,]  74   9      1          8        0   200
#> [469,]  75   6      0        180        0   150
#> [470,]  79  10      1          8        0   190
#> [471,]  77   3      0        180        0   110
#> [472,]  76  29      0         47        0    90
#> [473,]  73   8      1          1        1   162
#> [474,]  73  11      1          2        1   110
#> [475,]  74   2      0        180        0   100
#> [476,]  78   7      0          7        1   133
#> [477,]  78   8      1          6        1   110
#> [478,]  74   7      0          7        0   161
#> [479,]  76  13      1          1        1   170
#> [480,]  79   6      0        180        0   170
#> [481,]  80  10      1          6        1   147
#> [482,]  78  13      1          5        0   130
#> [483,]  75   5      0        119        1   150
#> [484,]  75  12      1          1        1   120
#> [485,]  78  15      0        180        1   270
#> [486,]  80   8      0          8        1   120
#> [487,]  75  13      1          6        0   150
#> [488,]  78  12      1          9        0   150
#> [489,]  78   2      1          1        0   130
#> [490,]  76   7      0         29        1   150
#> [491,]  77  24      0         24        1   160
#> [492,]  79   8      0         32        1   120
#> [493,]  80   6      0          6        1   150
#> [494,]  78   6      1          0        1   240
#> [495,]  76   3      1          0        1   140
#> [496,]  78  11      1          1        1   140
#> [497,]  79  11      0        180        0   160
#> [498,]  78  14      1          0        1   140
#> [499,]  78  11      1          8        1   118
#> [500,]  76   4      0          4        1   160
#> [501,]  79   4      0          4        1   125
#> [502,]  76  10      1          8        0   180
#> [503,]  76  12      1         10        1   127
#> [504,]  77   6      0          6        1   107
#> [505,]  75   2      1          1        1   204
#> [506,]  78  11      0        180        1   135
#> [507,]  77  31      1          3        1   161
#> [508,]  76   1      0          1        1    90
#> [509,]  77   7      0        180        1   170
#> [510,]  79   4      1          0        1   120
#> [511,]  81   1      0        180        0   120
#> [512,]  80  15      1         12        1   150
#> [513,]  80  40      1          0        1   138
#> [514,]  78   4      0         59        1   112
#> [515,]  76   7      0        161        0   151
#> [516,]  79  10      0         10        1   120
#> [517,]  80  15      1          0        1    90
#> [518,]  81   4      1          2        1   126
#> [519,]  79  28      0        164        0   100
#> [520,]  80   9      0        118        1   186
#> [521,]  79   1      0         37        1   140
#> [522,]  81   3      0        180        0   184
#> [523,]  78   7      0          7        1   147
#> [524,]  78  15      0         15        0   165
#> [525,]  80   5      1          1        1   108
#> [526,]  78  26      1          5        0   194
#> [527,]  76   1      0        166        0   131
#> [528,]  80   1      0          1        0   100
#> [529,]  78   3      1          1        1   152
#> [530,]  77  10      1          8        1   130
#> [531,]  77   5      0         85        0   188
#> [532,]  80   2      1          1        0   168
#> [533,]  80   6      1          0        1   119
#> [534,]  80   5      0          5        1   130
#> [535,]  82   1      1          0        1    82
#> [536,]  77   4      0        180        1    98
#> [537,]  81   1      0        108        0   129
#> [538,]  78  12      0        180        0   134
#> [539,]  79   1      0        125        0   193
#> [540,]  82  21      1          2        0   155
#> [541,]  84  22      1         10        0   180
#> [542,]  79   4      0          4        1   121
#> [543,]  80   6      0          6        1   110
#> [544,]  83   9      1          5        1   170
#> [545,]  82   5      0        180        0   110
#> [546,]  83   5      0        180        0   148
#> [547,]  79   7      1          6        0   130
#> [548,]  83   4      0        103        0    97
#> [549,]  81  11      1          8        0   160
#> [550,]  81   5      0        177        0    41
#> [551,]  80  11      1          8        0   170
#> [552,]  79   4      0          4        1   183
#> [553,]  79   1      0        180        1   170
#> [554,]  81  15      0        180        1   140
#> [555,]  80   7      1          0        1   146
#> [556,]  84   5      1          1        1    85
#> [557,]  81  20      1          9        0   170
#> [558,]  81  16      0         16        1   110
#> [559,]  80  11      1          8        0   110
#> [560,]  81   8      0        180        0   146
#> [561,]  80   8      1          7        0   160
#> [562,]  79   0      1          0        1    96
#> [563,]  85   4      0        180        0    90
#> [564,]  81   2      1          1        0   198
#> [565,]  84   4      0        167        0   198
#> [566,]  80   3      1          1        1   230
#> [567,]  82  23      1          0        0   110
#> [568,]  84   4      0          4        1    85
#> [569,]  84   1      0         38        1   205
#> [570,]  81   4      0         90        1   138
#> [571,]  79   9      1          8        0   150
#> [572,]  85   3      1          2        1   160
#> [573,]  80   2      1          0        1   130
#> [574,]  79   4      0          4        1    60
#> [575,]  80   6      0         71        1   189
#> [576,]  83   1      0          1        1   100
#> [577,]  82  19      0         19        0   120
#> [578,]  80  30      1         13        0   220
#> [579,]  83   3      0        114        0    98
#> [580,]  81  14      1         12        1   128
#> [581,]  83   2      0        154        0   130
#> [582,]  82   0      0          2        1   100
#> [583,]  83   1      0        180        0   160
#> [584,]  81   4      0          4        0   175
#> [585,]  84  15      1         13        1   110
#> [586,]  82  16      1          8        0   103
#> [587,]  82   5      1          0        1   146
#> [588,]  81   4      0          4        0   160
#> [589,]  86  12      0        180        1   120
#> [590,]  83  12      1          2        1   170
#> [591,]  81  19      1         14        0   120
#> [592,]  82   3      1          2        0   130
#> [593,]  82  15      1          0        0   183
#> [594,]  80   2      0         88        0   135
#> [595,]  86   8      0          8        1   132
#> [596,]  81  16      1          9        0   180
#> [597,]  84   6      0        165        0   145
#> [598,]  82   9      0        180        1   134
#> [599,]  84   3      0        180        1   120
#> [600,]  81  13      0        180        0   152
#> [601,]  85   3      0          3        1   118
#> [602,]  81   2      1          0        1   118
#> [603,]  82   1      0        180        1   193
#> [604,]  83   4      0          4        0   130
#> [605,]  86  12      1          0        1   132
#> [606,]  86   6      1          0        1   140
#> [607,]  84  16      0         70        1   150
#> [608,]  84   3      1          2        0   125
#> [609,]  83  10      1          0        1   190
#> [610,]  86   2      0        180        1   169
#> [611,]  88  14      1          3        1   130
#> [612,]  84   7      1          2        0   148
#> [613,]  86   4      0         38        1   122
#> [614,]  82   4      0          4        0   130
#> [615,]  86   6      0          6        1   117
#> [616,]  84  13      0         62        1   100
#> [617,]  88   4      0          4        0   100
#> [618,]  83  20      1          3        1   150
#> [619,]  88   4      0          4        1   115
#> [620,]  85  22      0         22        1   184
#> [621,]  83   9      0         65        1   150
#> [622,]  87   2      0        180        1   130
#> [623,]  88   3      0        115        0   110
#> [624,]  83   3      0          3        1   130
#> [625,]  87   8      0          8        1   157
#> [626,]  88   4      0          4        0    86
#> [627,]  89   4      0          4        1   153
#> [628,]  89   5      0        119        1   140
#> [629,]  87   6      0        180        1   110
#> [630,]  87   1      0          1        0   170
#> [631,]  84   8      0        180        1   119
#> [632,]  84   0      0        180        1   136
#> [633,]  89  10      0         46        1   170
#> [634,]  88   1      0          1        0   135
#> [635,]  86   4      0        180        1   145
#> [636,]  87   2      0        180        0   160
#> [637,]  91  10      0        145        0   135
#> [638,]  88   7      0         24        0   119
#> [639,]  90  11      1         10        1   186
#> [640,]  87   6      0        126        1   168
#> [641,]  86  10      0        180        1   137
#> [642,]  86   9      1          7        0   130
#> [643,]  90   4      1          0        0   121
#> [644,]  91   1      0          1        1    74
#> [645,]  87  43      0        178        1   130
#> [646,]  87   5      0         36        1   150
#> [647,]  90   5      1          0        1   125
#> [648,]  88   3      1          2        0   159
#> [649,]  89   3      1          1        1   160
#> [650,]  92   1      0          1        1   167
#> [651,]  88   5      0        158        0   100
#> [652,]  89  12      1          0        1   130
#> [653,]  89   2      0        168        0   118
#> [654,]  89  52      0         52        1   130
#> [655,]  89   4      0          4        1   159
#> [656,]  91   4      1          0        1   120
#> [657,]  90  19      1         11        1   129
#> [658,]  94   6      0         50        0    78
#> [659,]  90   1      0          1        1   118
#> [660,]  93   8      0        179        1   110
#> [661,]  91   1      0        180        0   158
#> [662,]  90  16      0         16        1   106
#> [663,]  96   3      0         12        1    97
#> [664,]  95   8      1          5        1   150
#> [665,]  94   3      0         26        1   144
#> [666,]  91  12      0         53        1   212
#> [667,]  91   7      0          7        0   135
#> [668,]  92   5      0         69        0   139
#> [669,]  93   4      0        180        1   135
#> [670,]  96   3      1          0        1   104
#> 
#> $y
#>   [1]   5.0+   2.0+   2.0+ 180.0+ 115.0  180.0+  12.0    5.0+ 180.0+ 180.0+
#>  [11] 180.0+ 180.0+ 180.0+   5.0+   2.0+ 180.0+ 180.0+   3.0  180.0+ 180.0+
#>  [21] 180.0+   2.0+ 180.0+ 155.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [31] 180.0+ 180.0+ 180.0+ 180.0+   6.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0+
#>  [41] 161.0+ 180.0+ 180.0+   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [51] 180.0+ 172.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0    1.0 
#>  [61] 179.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [71] 180.0+ 180.0+  36.0   88.0+ 180.0+   4.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [81] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  85.0  180.0+ 180.0+
#>  [91] 180.0+ 152.0+   6.0+ 180.0+ 180.0+  13.0+ 171.0+ 180.0+ 180.0+ 174.0+
#> [101] 180.0+ 180.0+ 180.0+ 175.0+ 180.0+   2.0  180.0+ 180.0+ 180.0+ 180.0+
#> [111] 180.0+ 180.0+  16.0+  15.0+ 180.0+ 180.0+ 180.0+  12.0+ 180.0+   8.0 
#> [121]   2.0    3.0+ 180.0+ 180.0+ 140.0  180.0+   1.0  165.0  180.0+ 180.0+
#> [131] 180.0+   8.0+ 180.0+ 180.0+   0.5  180.0+ 180.0+ 180.0+ 180.0+ 171.0+
#> [141]  15.0  166.0+ 180.0+   4.0+ 147.0+ 180.0+   5.0+ 180.0+   9.0+   1.0 
#> [151] 180.0+ 180.0+ 180.0+ 180.0+   2.0  180.0+ 180.0+ 180.0+ 180.0+  64.0 
#> [161] 180.0+ 180.0+ 180.0+   9.0+ 180.0+   0.5  171.0+ 180.0+   1.0  180.0+
#> [171] 180.0+  10.0+ 180.0+  45.0    3.0+  19.0  180.0+ 180.0+   9.0+ 172.0+
#> [181] 172.0+  24.0  180.0+   8.0  180.0+ 180.0+   1.0+ 180.0+ 180.0+ 180.0+
#> [191]  77.0    8.0+ 180.0+ 180.0+ 180.0+ 170.0   94.0    7.0  180.0+ 180.0+
#> [201] 180.0+ 180.0+ 180.0+  28.0    5.0   18.0  180.0+ 180.0+ 180.0+ 180.0+
#> [211] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  22.0  180.0+ 180.0+  84.0  180.0+
#> [221]   1.0    1.0  180.0+   4.0+   3.0+ 180.0+ 180.0+  12.0  180.0+ 180.0+
#> [231] 180.0+  36.0  180.0+   3.0+ 180.0+  12.0  180.0+ 180.0+ 180.0+ 180.0+
#> [241]  12.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  12.0    9.0  180.0+ 180.0+
#> [251] 180.0+   2.0+  18.0+ 180.0+ 180.0+ 180.0+   3.0+   2.0+ 103.0  180.0+
#> [261] 180.0+  13.0  166.0+   3.0    0.5+   3.0+ 180.0+ 175.0+ 180.0+ 180.0+
#> [271]   7.0+   8.0    5.0   16.0  180.0+   1.0  180.0+ 180.0+ 123.0+   1.0+
#> [281] 180.0+  15.0+   4.0+ 180.0+ 180.0+ 174.0+ 180.0+ 180.0+ 180.0+   8.0+
#> [291]   3.0  180.0+  10.0  180.0+   6.0  180.0+  19.0+  12.0  179.0+ 180.0+
#> [301] 180.0+  11.0+   0.5   18.0    7.0+ 180.0+ 180.0+ 180.0+ 180.0+  21.0+
#> [311]   1.0   18.0+   4.0    7.0+ 150.0  180.0+ 180.0+ 180.0+ 171.0  180.0+
#> [321]   6.0  180.0+   0.5  180.0+ 180.0+ 180.0+ 180.0+   2.0   45.0    5.0+
#> [331] 103.0    3.0+ 180.0+ 180.0+  97.0  180.0+   8.0+ 180.0+ 180.0+  18.0 
#> [341] 172.0+ 180.0+ 180.0+   7.0  180.0+   7.0   13.0+ 123.0  180.0+  51.0 
#> [351]  19.0    1.0   60.0   76.0  132.0   10.0+ 162.0    7.0+   7.0+   9.0 
#> [361]  12.0  180.0+ 180.0+ 180.0+ 173.0+ 180.0+ 180.0+ 180.0+ 180.0+  28.0 
#> [371] 180.0+  16.0+ 180.0+   6.0  180.0+ 180.0+  15.0    3.0+ 180.0+ 180.0+
#> [381] 180.0+   2.0    3.0+ 180.0+  20.0  180.0+  20.0    0.5  180.0+ 180.0+
#> [391]   8.0    3.0   12.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   3.0 
#> [401] 175.0   14.0+ 180.0+  10.0+ 180.0+ 180.0+   8.0+ 179.0+   1.0  180.0+
#> [411] 180.0+  15.0  180.0+  10.0  180.0+   4.0+ 180.0+  10.0   57.0  180.0+
#> [421]  11.0    3.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 177.0+ 180.0+
#> [431]   4.0+   7.0   15.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  85.0 
#> [441] 180.0+  17.0+   4.0    7.0    0.5   12.0  180.0+ 180.0+  46.0  180.0+
#> [451]   4.0  180.0+ 180.0+  12.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0 
#> [461] 180.0+ 180.0+  12.0  180.0+   7.0+  79.0    3.0  168.0+ 180.0+ 180.0+
#> [471] 180.0+  47.0  180.0+  11.0  180.0+   7.0    8.0+   7.0  180.0+ 180.0+
#> [481]  10.0  172.0  119.0   12.0  180.0+   8.0  180.0+ 180.0+ 180.0+  29.0 
#> [491]  24.0   32.0    6.0  180.0+   3.0+ 180.0+ 180.0+ 180.0+  11.0    4.0 
#> [501]   4.0   10.0+ 180.0+   6.0    2.0+ 180.0+ 171.0    1.0  180.0+ 138.0 
#> [511] 180.0+ 180.0+  40.0   59.0  161.0   10.0+ 180.0+  93.0  164.0  118.0 
#> [521]  37.0  180.0+   7.0+  15.0+   5.0+ 171.0+ 166.0+   1.0    3.0+  10.0 
#> [531]  85.0   10.0    6.0    5.0    1.0  180.0+ 108.0  180.0+ 125.0  180.0+
#> [541] 180.0+   4.0    6.0    9.0+ 180.0+ 180.0+ 180.0+ 103.0  180.0+ 177.0+
#> [551] 169.0    4.0  180.0+ 180.0+   7.0+ 180.0+  20.0   16.0  180.0+ 180.0+
#> [561] 180.0+   0.5  180.0+ 180.0+ 167.0    3.0+  62.0    4.0   38.0   90.0 
#> [571] 180.0+ 180.0+ 180.0+   4.0   71.0    1.0   19.0   30.0  114.0  180.0+
#> [581] 154.0    2.0  180.0+   4.0+ 180.0+  16.0+   5.0+   4.0+ 180.0+  77.0 
#> [591] 180.0+   3.0   83.0   88.0    8.0  180.0+ 165.0  180.0+ 180.0+ 180.0+
#> [601]   3.0+ 180.0+ 180.0+   4.0+ 180.0+   6.0   70.0  180.0+ 180.0+ 180.0+
#> [611]  14.0  180.0+  38.0    4.0    6.0+  62.0    4.0+  20.0    4.0   22.0 
#> [621]  65.0  180.0+ 115.0    3.0+   8.0+   4.0    4.0  119.0  180.0+   1.0+
#> [631] 180.0+ 180.0+  46.0    1.0+ 180.0+ 180.0+ 145.0   24.0   11.0  126.0 
#> [641] 180.0+ 180.0+   4.0    1.0  178.0+  36.0   89.0   75.0    3.0+   1.0 
#> [651] 158.0  180.0+ 168.0   52.0    4.0    4.0  180.0+  50.0    1.0+ 179.0+
#> [661] 180.0+  16.0   12.0    8.0   26.0   53.0    7.0+  69.0  180.0+   3.0 
#> 
#> $weights
#> NULL
#> 
#> $offset
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Warning: Multiple lambdas have been fit. Lambda will be set to 0.01 (see parameter 's').

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8375425