GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.coxnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.coxnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Caution: This learner is different to learners calling glmnet::cv.glmnet() in that it does not use the internal optimization of parameter lambda. Instead, lambda needs to be tuned by the user (e.g., via mlr3tuning). When lambda is tuned, the glmnet will be trained for each tuning iteration. While fitting the whole path of lambdas would be more efficient, as is done by default in glmnet::glmnet(), tuning/selecting the parameter at prediction time (using parameter s) is currently not supported in mlr3 (at least not in efficient manner). Tuning the s parameter is, therefore, currently discouraged.

When the data are i.i.d. and efficiency is key, we recommend using the respective auto-tuning counterpart in mlr_learners_surv.cv_glmnet(). However, in some situations this is not applicable, usually when data are imbalanced or not i.i.d. (longitudinal, time-series) and tuning requires custom resampling strategies (blocked design, stratification).

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.coxnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exact	logical	FALSE	TRUE, FALSE	-

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	TRUE	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	0.01		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.logistic	character	Newton	Newton, modified.Newton	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvGlmnet

Methods

Public methods

LearnerSurvGlmnet$new()

LearnerSurvGlmnet$selected_features()

LearnerSurvGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("surv.glmnet")
print(learner)
#> 
#> ── <LearnerSurvGlmnet> (surv.glmnet): Regularized Generalized Linear Model ─────
#> • Model: -
#> • Parameters: use_pred_offset=TRUE
#> • Packages: mlr3, mlr3proba, mlr3extralearners, and glmnet
#> • Predict Types: [crank], distr, and lp
#> • Feature Types: logical, integer, and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: offset, selected_features, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("grace")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#>    Df  %Dev   Lambda
#> 1   0  0.00 0.176300
#> 2   2  0.77 0.160700
#> 3   2  1.71 0.146400
#> 4   2  2.53 0.133400
#> 5   3  5.22 0.121500
#> 6   3  7.84 0.110700
#> 7   3  9.81 0.100900
#> 8   3 11.34 0.091930
#> 9   3 12.57 0.083760
#> 10  4 13.63 0.076320
#> 11  4 14.53 0.069540
#> 12  4 15.28 0.063360
#> 13  4 15.90 0.057740
#> 14  4 16.43 0.052610
#> 15  4 16.87 0.047930
#> 16  5 17.25 0.043670
#> 17  5 17.58 0.039790
#> 18  6 17.87 0.036260
#> 19  6 18.13 0.033040
#> 20  6 18.36 0.030100
#> 21  6 18.55 0.027430
#> 22  6 18.71 0.024990
#> 23  6 18.84 0.022770
#> 24  6 18.96 0.020750
#> 25  6 19.05 0.018910
#> 26  6 19.14 0.017230
#> 27  6 19.20 0.015700
#> 28  6 19.26 0.014300
#> 29  6 19.31 0.013030
#> 30  6 19.35 0.011870
#> 31  6 19.39 0.010820
#> 32  6 19.42 0.009857
#> 33  6 19.44 0.008982
#> 34  6 19.46 0.008184
#> 35  6 19.48 0.007457
#> 36  6 19.49 0.006794
#> 37  6 19.51 0.006191
#> 38  6 19.52 0.005641
#> 39  6 19.52 0.005140
#> 40  6 19.53 0.004683
#> 41  6 19.54 0.004267
#> 42  6 19.54 0.003888
#> 43  6 19.55 0.003543
#> 44  6 19.55 0.003228
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  28   9      0        180        1   107
#>   [2,]  32   5      1          0        1   121
#>   [3,]  35  10      1          9        0   106
#>   [4,]  34   5      0          5        0   120
#>   [5,]  35   2      1          1        1   112
#>   [6,]  36   1      0        180        1   155
#>   [7,]  35   0      0        180        1   119
#>   [8,]  38  12      1          8        1   120
#>   [9,]  36   5      1          0        1   115
#>  [10,]  33   6      1          1        1   115
#>  [11,]  38  12      1         11        1    92
#>  [12,]  42   2      0        180        1   100
#>  [13,]  38   5      1          3        0   125
#>  [14,]  42   2      0          2        0   140
#>  [15,]  40   6      0        180        1   138
#>  [16,]  40  11      1         10        1   120
#>  [17,]  42   2      0        180        0   100
#>  [18,]  41   2      1          1        0   166
#>  [19,]  43   4      1          0        1   130
#>  [20,]  42   4      0        180        0   162
#>  [21,]  42   2      0        180        1   124
#>  [22,]  44   5      1          1        0   170
#>  [23,]  45   3      0        180        1   154
#>  [24,]  44   3      0        180        0   141
#>  [25,]  45   9      1          7        0   110
#>  [26,]  44   2      1          1        1   150
#>  [27,]  43   2      0        180        1   140
#>  [28,]  45   2      0        180        1   140
#>  [29,]  46  15      0        180        0   120
#>  [30,]  46   2      1          1        0   126
#>  [31,]  47   4      1          3        0   118
#>  [32,]  48  15      0        180        1   160
#>  [33,]  45   3      0        150        0   130
#>  [34,]  44   3      1          0        1   180
#>  [35,]  46   7      1          2        0   166
#>  [36,]  47   6      1          0        1   116
#>  [37,]  44   0      1          0        1    96
#>  [38,]  43   3      1          0        1   124
#>  [39,]  45   8      1          0        1   117
#>  [40,]  45   5      0          5        0   141
#>  [41,]  46   2      1          1        1   122
#>  [42,]  46   6      1          0        1   100
#>  [43,]  44   4      1          0        1   114
#>  [44,]  47   2      0        180        0   108
#>  [45,]  44   9      1          8        1   135
#>  [46,]  46   5      1          3        0   130
#>  [47,]  46   4      0        180        1   121
#>  [48,]  44   2      0        180        0   142
#>  [49,]  45   9      1          0        1   145
#>  [50,]  47   3      1          1        1   120
#>  [51,]  48   3      0        180        0   154
#>  [52,]  48  12      1         11        0   200
#>  [53,]  47   5      1          3        1   130
#>  [54,]  46   3      1          0        1   119
#>  [55,]  49   4      0        180        0   117
#>  [56,]  47  10      0         10        1   140
#>  [57,]  48   2      1          0        0   184
#>  [58,]  50   6      1          2        1   140
#>  [59,]  49   7      1          7        1   110
#>  [60,]  46   9      1          9        1   122
#>  [61,]  50   7      0        180        1   110
#>  [62,]  49   2      0          2        0   105
#>  [63,]  51   1      0          1        1   145
#>  [64,]  49  15      1         11        1   160
#>  [65,]  47   2      0        180        0   150
#>  [66,]  49  23      0        179        1   112
#>  [67,]  46   6      1          0        1   156
#>  [68,]  52   2      0        180        1   170
#>  [69,]  50   4      0          4        1   100
#>  [70,]  51   3      1          2        0   113
#>  [71,]  50   1      1          0        0   150
#>  [72,]  49   7      1          4        1    90
#>  [73,]  47   8      0        180        0   160
#>  [74,]  50   4      1          1        0   150
#>  [75,]  48   7      1          0        1   110
#>  [76,]  53   8      0         36        1   160
#>  [77,]  48  17      1         10        0   111
#>  [78,]  47   2      1          1        0   110
#>  [79,]  52   4      1          4        0   152
#>  [80,]  49   9      1          3        0   102
#>  [81,]  49  15      0        180        1   160
#>  [82,]  53   5      0        180        1   140
#>  [83,]  54  17      1         12        1   102
#>  [84,]  53   5      0         77        0   159
#>  [85,]  54   6      1          3        0   129
#>  [86,]  50   2      0          5        1   106
#>  [87,]  53   8      1          7        0   160
#>  [88,]  48   3      1          2        0   150
#>  [89,]  51  25      1          1        0   202
#>  [90,]  53   4      0          4        0   140
#>  [91,]  52  14      1          7        1   200
#>  [92,]  53   4      1          0        1   156
#>  [93,]  51  13      0         99        1   160
#>  [94,]  54   9      1          0        1   138
#>  [95,]  54  23      1         10        0   131
#>  [96,]  52   7      1          2        0   154
#>  [97,]  55   6      1          2        1   114
#>  [98,]  54   9      1          1        0   130
#>  [99,]  55   4      1          2        0   150
#> [100,]  51  13      1         11        0   145
#> [101,]  50   5      1          4        1   150
#> [102,]  54   4      1          0        1   121
#> [103,]  52   4      0        180        0   183
#> [104,]  50   3      0        174        1   153
#> [105,]  49   6      1          0        1   130
#> [106,]  49   1      0          1        1   110
#> [107,]  50   7      1          1        0   156
#> [108,]  53   9      0          9        1    95
#> [109,]  53   8      1          0        1   130
#> [110,]  50   7      1          0        1   127
#> [111,]  55   1      0        180        0   127
#> [112,]  55   2      0          2        0   145
#> [113,]  56   2      0        180        0   132
#> [114,]  52   8      0        180        0   119
#> [115,]  53  18      1          9        1   150
#> [116,]  54   3      0        180        1   180
#> [117,]  55   6      0        180        0   170
#> [118,]  52  16      0         16        0   152
#> [119,]  53  10      1          9        0   172
#> [120,]  53  15      0         15        1    90
#> [121,]  55   6      1          5        1   138
#> [122,]  54  12      1          0        1   190
#> [123,]  54   3      0        180        0   128
#> [124,]  56   3      0          8        1   139
#> [125,]  57   3      0          3        0   120
#> [126,]  54   7      1          2        0   129
#> [127,]  54   2      1          1        0   135
#> [128,]  52   9      1          3        0   170
#> [129,]  54   2      1          1        1   176
#> [130,]  57   5      1          3        1   138
#> [131,]  57   1      0        180        1   156
#> [132,]  57   1      0          1        1   100
#> [133,]  56   4      1          0        1   140
#> [134,]  55  11      1          7        0   104
#> [135,]  56  14      1         11        0   130
#> [136,]  53   3      1          0        1   200
#> [137,]  54   5      0        180        1   108
#> [138,]  55   3      1          1        1   156
#> [139,]  57   0      0          0        1   150
#> [140,]  59   3      1          1        0   172
#> [141,]  57   4      0        180        1   119
#> [142,]  58   6      1          0        1    90
#> [143,]  53  15      1         10        1   130
#> [144,]  54  17      1          8        1   227
#> [145,]  55   9      1          2        1   147
#> [146,]  56   5      0          5        1   150
#> [147,]  57   4      1          2        1   185
#> [148,]  53   4      0        147        1   145
#> [149,]  53   7      1          0        1   120
#> [150,]  55   3      1          2        0   140
#> [151,]  55   5      0          5        1   131
#> [152,]  56   4      0          4        0   164
#> [153,]  58   9      1          0        1   180
#> [154,]  58   1      1          1        1   200
#> [155,]  55   2      0          2        0   106
#> [156,]  59   9      1          1        1   125
#> [157,]  57   1      0        180        0   148
#> [158,]  59   3      0        180        0   120
#> [159,]  57   2      0          2        1   120
#> [160,]  57   5      0        180        1   130
#> [161,]  55   5      1          0        1   160
#> [162,]  57  10      1          9        0   103
#> [163,]  59   5      0        180        1   155
#> [164,]  58  26      1          0        1   189
#> [165,]  58   4      1          3        0   120
#> [166,]  60   0      1          0        1    80
#> [167,]  59   2      1          1        0   140
#> [168,]  58   8      0        161        1   140
#> [169,]  61   4      1          3        0   151
#> [170,]  61   9      1          8        0   150
#> [171,]  61   3      1          2        1   102
#> [172,]  61  20      1         13        0   130
#> [173,]  57   2      1          0        1   116
#> [174,]  58  10      0         10        1   150
#> [175,]  57   4      1          3        0   138
#> [176,]  57  11      0        180        1   150
#> [177,]  56  14      0         45        0   130
#> [178,]  58  19      1         13        1   140
#> [179,]  56  13      1          6        1   158
#> [180,]  56  18      1         11        1   165
#> [181,]  55   9      1          7        1   135
#> [182,]  61   4      1          0        1   115
#> [183,]  61  13      1         12        1   130
#> [184,]  59  11      1          8        1   190
#> [185,]  59   5      1          2        0   182
#> [186,]  58   5      1          1        1   135
#> [187,]  59  10      0        180        0   160
#> [188,]  61   8      0         77        0   120
#> [189,]  61  13      0         13        0   210
#> [190,]  62  10      1          0        1   153
#> [191,]  62   7      1          2        1   180
#> [192,]  57   3      1          0        0   100
#> [193,]  61  18      0        170        0   140
#> [194,]  58   8      1          3        1   150
#> [195,]  61   7      0          7        1   150
#> [196,]  60   7      0          7        0   147
#> [197,]  61   6      0          6        0   134
#> [198,]  62   4      1          0        0   160
#> [199,]  60  17      1          8        1   140
#> [200,]  58   3      1          0        1   146
#> [201,]  62   4      1          3        0   173
#> [202,]  58   2      0         30        0   202
#> [203,]  59   1      0        180        0   155
#> [204,]  61   5      0          5        1   110
#> [205,]  57  18      1          9        1    93
#> [206,]  61   5      0          5        1   160
#> [207,]  58  11      1          9        0   179
#> [208,]  62  17      1         10        1   180
#> [209,]  63   3      1          1        0   180
#> [210,]  61   7      0        180        0   135
#> [211,]  63   4      1          3        0   222
#> [212,]  62   3      0        180        1   105
#> [213,]  63  15      1         10        1   126
#> [214,]  64   4      0        180        0   130
#> [215,]  63   4      1          1        0   155
#> [216,]  60  18      1         13        0   132
#> [217,]  59   8      0        180        1   140
#> [218,]  58   9      1          9        0   110
#> [219,]  58   2      0        180        0   127
#> [220,]  60   7      0          7        0   140
#> [221,]  59   5      1          1        0   148
#> [222,]  65  13      0        180        1   100
#> [223,]  63   1      0          1        0   130
#> [224,]  62   6      0        180        0   170
#> [225,]  61  15      1         13        0   170
#> [226,]  59   4      0          4        0   149
#> [227,]  64  10      1          9        0   160
#> [228,]  59  10      0        180        1   130
#> [229,]  60   8      0         17        1   130
#> [230,]  61   6      1          1        1   117
#> [231,]  64  12      1         11        0   160
#> [232,]  64   6      1          0        1   140
#> [233,]  63  10      1          0        1   148
#> [234,]  63  14      1          9        0   123
#> [235,]  65  36      1         11        0   140
#> [236,]  66   3      1          1        0   127
#> [237,]  61  10      1          2        1   194
#> [238,]  64  32      1          9        1   160
#> [239,]  63  12      1          9        0   114
#> [240,]  63   7      0        180        0   120
#> [241,]  65   8      1          0        0   168
#> [242,]  65  10      1          8        1   120
#> [243,]  64  21      1         10        0   190
#> [244,]  61  12      1         11        0   154
#> [245,]  64   9      0        180        0   150
#> [246,]  61   4      0        180        1   113
#> [247,]  63  16      1          7        1   110
#> [248,]  66   6      1          1        1   130
#> [249,]  65   6      0          9        0   112
#> [250,]  63   5      1          4        0   170
#> [251,]  63   2      1          1        0   180
#> [252,]  62  13      1         11        0   180
#> [253,]  64   2      0          2        0   201
#> [254,]  66  18      1          5        0   142
#> [255,]  62   9      0        180        0   145
#> [256,]  61  15      1         10        0   130
#> [257,]  63   9      1          8        1   160
#> [258,]  63   3      1          2        0   120
#> [259,]  63   2      1          0        0   140
#> [260,]  67   6      0        180        1   170
#> [261,]  65  15      1         11        1   160
#> [262,]  64  13      1         12        1   150
#> [263,]  66   7      1          0        1   115
#> [264,]  66  13      1          0        0   118
#> [265,]  64  14      1         13        1   150
#> [266,]  64   0      0          0        1   148
#> [267,]  67   4      1          3        0   130
#> [268,]  66   6      1          0        1   140
#> [269,]  65   2      1          1        1   170
#> [270,]  68   1      0        180        1   166
#> [271,]  63   7      1          0        0   162
#> [272,]  67   8      1          1        1   130
#> [273,]  68   5      0          5        1    90
#> [274,]  63  10      0         16        1   160
#> [275,]  66  14      0        180        0   130
#> [276,]  64   1      0          1        1   120
#> [277,]  68  18      0        180        1   260
#> [278,]  63   8      1          1        1   162
#> [279,]  63   1      1          0        1   155
#> [280,]  67  11      0         11        0   150
#> [281,]  68  11      0        180        0   160
#> [282,]  68  14      0         79        0   172
#> [283,]  66  12      1         10        1   150
#> [284,]  65  15      1         12        1   150
#> [285,]  66  11      1          0        0   100
#> [286,]  69  12      0         15        1   140
#> [287,]  63   2      0        180        0   150
#> [288,]  65  11      1          6        0   130
#> [289,]  69   6      0        180        1   100
#> [290,]  63   8      0        180        1   120
#> [291,]  68  14      1         13        1   140
#> [292,]  65   8      1          0        1    90
#> [293,]  67   1      0        180        1   160
#> [294,]  65   1      1          0        0   133
#> [295,]  63   2      1          0        0    99
#> [296,]  67   2      0        180        0   184
#> [297,]  65   6      0          6        0    80
#> [298,]  65  10      1          1        1   148
#> [299,]  66  19      1         12        1   150
#> [300,]  67  12      1         12        0   160
#> [301,]  69   6      0         99        1   140
#> [302,]  64   4      0        179        0   160
#> [303,]  66   4      0        180        1   130
#> [304,]  70  15      1         12        1   132
#> [305,]  64  11      0         11        0   125
#> [306,]  64   4      0        180        1   140
#> [307,]  64   0      1          0        1   118
#> [308,]  67   2      0         18        0   131
#> [309,]  66   7      1          5        1   131
#> [310,]  66   4      0        180        0   177
#> [311,]  68   4      1          0        1   160
#> [312,]  69   4      1          3        1   150
#> [313,]  65  13      1         12        1   130
#> [314,]  69  17      1         10        0   140
#> [315,]  69   8      0         93        0   140
#> [316,]  64  21      0         21        1   155
#> [317,]  65   6      0        101        1   115
#> [318,]  71  20      1          0        1   160
#> [319,]  67   2      0        180        0   128
#> [320,]  66   1      1          1        1   165
#> [321,]  68  18      1         14        1   170
#> [322,]  69   0      0          0        1   148
#> [323,]  65   2      0        180        0   130
#> [324,]  69   3      1          2        0   151
#> [325,]  67  14      1         13        0   130
#> [326,]  71   7      0          7        0   230
#> [327,]  66   2      0          2        1   228
#> [328,]  71   6      0         45        1   158
#> [329,]  69   5      0          5        1   142
#> [330,]  71   3      0        103        0   133
#> [331,]  70  22      1         13        0   103
#> [332,]  67   1      0         36        1   104
#> [333,]  67   5      0          5        0   130
#> [334,]  68   6      0        180        0   145
#> [335,]  69   8      1          5        1   195
#> [336,]  69   6      1          4        1   174
#> [337,]  72   3      1          0        1   132
#> [338,]  72   7      0          7        1   110
#> [339,]  69   8      1          7        1   108
#> [340,]  66   2      1          1        0   123
#> [341,]  69  19      0        180        0   130
#> [342,]  68  18      0         18        1   100
#> [343,]  66   2      0        180        0   130
#> [344,]  68   2      0          7        1   130
#> [345,]  70   3      0        123        0   130
#> [346,]  72   5      1          4        0   170
#> [347,]  68   3      0         19        0   135
#> [348,]  67  12      1          8        0   120
#> [349,]  67   1      0          1        1    60
#> [350,]  69   5      0         76        0   120
#> [351,]  67   8      1          0        1   130
#> [352,]  68  10      1          8        1   160
#> [353,]  66  24      1         13        0   130
#> [354,]  70  35      1          0        1   105
#> [355,]  72  30      1          0        1   145
#> [356,]  68   7      1          2        0   135
#> [357,]  73  20      1          0        1   170
#> [358,]  71   6      0          9        0   120
#> [359,]  72  12      1         10        0   170
#> [360,]  67   8      0        180        1   170
#> [361,]  67   9      0        180        0   158
#> [362,]  73  13      0        152        1   130
#> [363,]  67   4      1          1        0   134
#> [364,]  72   6      1          5        0   115
#> [365,]  68  23      0        180        1   220
#> [366,]  70   3      0        180        0   121
#> [367,]  69   3      0        180        0   220
#> [368,]  71   3      1          2        0   150
#> [369,]  68   4      1          3        0   210
#> [370,]  72   5      0         28        0   120
#> [371,]  71   5      0        180        0   191
#> [372,]  73   6      0        180        1   117
#> [373,]  69   8      1          1        0   164
#> [374,]  68   7      0        180        1   130
#> [375,]  72  16      1          1        1   130
#> [376,]  70   4      0        180        0   180
#> [377,]  69   1      1          0        0   155
#> [378,]  71   2      1          0        1   180
#> [379,]  68  15      1         13        1   130
#> [380,]  70   3      0          3        1   159
#> [381,]  70  13      1          9        0   100
#> [382,]  72   6      0        180        1   130
#> [383,]  74   8      1          0        1    85
#> [384,]  73   4      0        180        1   154
#> [385,]  71   3      1          1        0   150
#> [386,]  71  15      1         11        0   165
#> [387,]  68   9      0        180        1   120
#> [388,]  74   0      1          0        1    90
#> [389,]  73   3      1          0        1   136
#> [390,]  70   5      1          0        1   190
#> [391,]  71  17      1         11        0   160
#> [392,]  73  10      1          8        0   106
#> [393,]  69  12      1          1        1   149
#> [394,]  70  26      1         11        1   120
#> [395,]  74   4      0          4        0   120
#> [396,]  73   6      0        180        0   110
#> [397,]  71   7      1          2        0   143
#> [398,]  72   8      1          0        1   140
#> [399,]  71  13      1          8        0   121
#> [400,]  69   2      1          1        1    80
#> [401,]  70   4      1          0        1   140
#> [402,]  71  14      1         13        1   170
#> [403,]  72  10      1          8        1   153
#> [404,]  69   7      0        180        1   144
#> [405,]  71  10      1          9        1   120
#> [406,]  75   2      1          1        0   145
#> [407,]  72  10      1          9        1   160
#> [408,]  73  10      1         10        1   120
#> [409,]  74  15      1          9        1   179
#> [410,]  71   2      0         10        1   112
#> [411,]  73   1      0          1        1    80
#> [412,]  75  13      1          1        1   130
#> [413,]  71  11      1          8        0   110
#> [414,]  71   4      0          4        0   134
#> [415,]  72  15      1         12        1   120
#> [416,]  73  10      1          8        0   120
#> [417,]  72   7      0         57        1   145
#> [418,]  73  10      0        180        0   162
#> [419,]  72  11      0         11        1   140
#> [420,]  70   3      0          3        0   150
#> [421,]  73   5      1          3        1   112
#> [422,]  76  25      1         12        1   170
#> [423,]  72   2      0        180        0   120
#> [424,]  75   1      0        180        1   140
#> [425,]  72   4      1          0        1   197
#> [426,]  71   3      1          0        0   144
#> [427,]  73   5      0        180        0   126
#> [428,]  73   4      0        180        0   124
#> [429,]  74  34      1          8        1   233
#> [430,]  76   3      1          0        1   120
#> [431,]  71  32      1         12        1   107
#> [432,]  72   5      0        180        0   154
#> [433,]  76   5      0          5        1   130
#> [434,]  77  11      0         11        1   150
#> [435,]  77   4      0          4        0   185
#> [436,]  75   3      1          1        0   180
#> [437,]  72   7      1          2        0   142
#> [438,]  73  15      0         15        1   160
#> [439,]  73  10      1         10        0   124
#> [440,]  74   7      0        180        1   150
#> [441,]  76   1      0        180        0   114
#> [442,]  74   2      1          1        0   140
#> [443,]  76   8      1          0        1   141
#> [444,]  74  19      1          4        1   200
#> [445,]  73   6      0          6        1   114
#> [446,]  74   2      0        180        0   190
#> [447,]  72   4      0         85        1   120
#> [448,]  76  17      1          0        1   200
#> [449,]  76  13      1         10        0   110
#> [450,]  75   4      1          0        1   122
#> [451,]  75   0      0          0        1   130
#> [452,]  73  13      1         11        0   195
#> [453,]  74   8      1          0        1   105
#> [454,]  76  13      1          8        1   148
#> [455,]  74   6      0        180        0   160
#> [456,]  76   4      0          4        1   155
#> [457,]  75   1      0          1        1   125
#> [458,]  73   1      0         52        1   105
#> [459,]  73   0      0        180        0   156
#> [460,]  72   5      0        180        0   120
#> [461,]  78  12      1         11        1   160
#> [462,]  76   5      0        180        0   185
#> [463,]  74   8      1          8        1   170
#> [464,]  73  33      1         12        1   175
#> [465,]  77  12      1          9        1   100
#> [466,]  73  10      1          9        0   146
#> [467,]  77   1      1          0        1    90
#> [468,]  78   5      1          0        1   170
#> [469,]  74   6      0         79        1   140
#> [470,]  74   9      1          8        0   200
#> [471,]  79  10      1          8        0   190
#> [472,]  77   3      0        180        0   110
#> [473,]  76  29      0         47        0    90
#> [474,]  73   8      1          1        1   162
#> [475,]  73  11      1          2        1   110
#> [476,]  74   2      0        180        0   100
#> [477,]  78   8      1          6        1   110
#> [478,]  74   7      0          7        0   161
#> [479,]  78  32      1          9        1   198
#> [480,]  79   6      0        180        0   170
#> [481,]  78   0      0        180        1   212
#> [482,]  78  13      1          5        0   130
#> [483,]  75  12      1          1        1   120
#> [484,]  80   8      0          8        1   120
#> [485,]  75  13      1          6        0   150
#> [486,]  74  10      1          8        0   135
#> [487,]  79   4      0         80        0   145
#> [488,]  78  12      1          9        0   150
#> [489,]  78  10      0        180        1   130
#> [490,]  76  11      1          0        0   120
#> [491,]  75  11      1          4        0   162
#> [492,]  75   3      0          3        0     0
#> [493,]  76   7      0         29        1   150
#> [494,]  77  24      0         24        1   160
#> [495,]  79   8      0         32        1   120
#> [496,]  80   9      0         23        1   128
#> [497,]  80   6      0          6        1   150
#> [498,]  76   3      1          0        1   140
#> [499,]  79  11      0        180        0   160
#> [500,]  79   2      1          0        1   121
#> [501,]  78  14      1          0        1   140
#> [502,]  81   1      0          1        0   130
#> [503,]  78  11      1          8        1   118
#> [504,]  79   4      0          4        1   125
#> [505,]  76  12      1         10        1   127
#> [506,]  80   3      1          0        1   120
#> [507,]  75   2      1          1        1   204
#> [508,]  78  11      0        180        1   135
#> [509,]  76   1      0          1        1    90
#> [510,]  77   7      0        180        1   170
#> [511,]  77   6      0          6        1   144
#> [512,]  79   4      1          0        1   120
#> [513,]  80  15      1         12        1   150
#> [514,]  77   9      1          4        0   141
#> [515,]  82   5      0          8        1   120
#> [516,]  80  40      1          0        1   138
#> [517,]  78   4      0         59        1   112
#> [518,]  80  17      1         12        0   100
#> [519,]  79  10      0         10        1   120
#> [520,]  80  15      1          0        1    90
#> [521,]  81   4      1          2        1   126
#> [522,]  80   9      0        118        1   186
#> [523,]  80   6      0        173        1   160
#> [524,]  81   3      0        180        0   184
#> [525,]  81   2      0        175        0   172
#> [526,]  77  13      1          0        1   190
#> [527,]  78  15      0         15        0   165
#> [528,]  78   4      0        180        0   175
#> [529,]  78  26      1          5        0   194
#> [530,]  76   1      0        166        0   131
#> [531,]  78  20      1          0        1   109
#> [532,]  80   1      0          1        0   100
#> [533,]  78   3      1          1        1   152
#> [534,]  82   3      1          1        1   144
#> [535,]  77   5      0         85        0   188
#> [536,]  80   2      1          1        0   168
#> [537,]  80   6      1          0        1   119
#> [538,]  78   2      0        180        0   148
#> [539,]  80   5      0          5        1   130
#> [540,]  82   1      1          0        1    82
#> [541,]  81   1      0        108        0   129
#> [542,]  78  12      0        180        0   134
#> [543,]  79   1      0        125        0   193
#> [544,]  82  21      1          2        0   155
#> [545,]  80   6      0          6        1   110
#> [546,]  83   9      1          5        1   170
#> [547,]  82   5      0        180        0   110
#> [548,]  79   7      1          6        0   130
#> [549,]  81  11      1          8        0   160
#> [550,]  81   5      0        177        0    41
#> [551,]  78  23      1         10        1   145
#> [552,]  79   4      0          4        1   183
#> [553,]  78   9      1          4        1   120
#> [554,]  82   8      1          1        0   128
#> [555,]  79   1      0        180        1   170
#> [556,]  81  15      0        180        1   140
#> [557,]  80   7      1          0        1   146
#> [558,]  84   5      1          1        1    85
#> [559,]  81  20      1          9        0   170
#> [560,]  83   8      0          8        0   115
#> [561,]  81  16      0         16        1   110
#> [562,]  80   6      1          0        1   150
#> [563,]  80  11      1          8        0   110
#> [564,]  81   8      0        180        0   146
#> [565,]  79   7      0        177        0   197
#> [566,]  85   4      0        180        0    90
#> [567,]  81   2      1          1        0   198
#> [568,]  83   2      0          2        1   155
#> [569,]  82   6      0        128        1   100
#> [570,]  84   5      0        180        1   203
#> [571,]  84   4      0          4        1    85
#> [572,]  81   1      0          1        1   150
#> [573,]  84   1      0         38        1   205
#> [574,]  79   9      1          8        0   150
#> [575,]  85   3      1          2        1   160
#> [576,]  80  13      1          8        1   140
#> [577,]  80   2      1          0        1   130
#> [578,]  79   4      0          4        1    60
#> [579,]  80   6      0         71        1   189
#> [580,]  83   1      0          1        1   100
#> [581,]  82  19      0         19        0   120
#> [582,]  80  30      1         13        0   220
#> [583,]  83   9      0        180        0   198
#> [584,]  79  14      1          0        0   110
#> [585,]  83   3      0        114        0    98
#> [586,]  81  14      1         12        1   128
#> [587,]  82   0      0          2        1   100
#> [588,]  83   1      0        180        0   160
#> [589,]  81   4      0          4        0   175
#> [590,]  84  15      1         13        1   110
#> [591,]  81   1      0          1        1   145
#> [592,]  82   5      1          0        1   146
#> [593,]  81   4      0          4        0   160
#> [594,]  86  12      0        180        1   120
#> [595,]  82   3      1          2        0   130
#> [596,]  82  15      1          0        0   183
#> [597,]  80   2      0         88        0   135
#> [598,]  83   7      0        126        0   135
#> [599,]  86   8      0          8        1   132
#> [600,]  84   6      0        165        0   145
#> [601,]  86   3      0          3        1   140
#> [602,]  82   9      0        180        1   134
#> [603,]  84   3      0        180        1   120
#> [604,]  81  13      0        180        0   152
#> [605,]  81   2      1          0        1   118
#> [606,]  83   9      0        180        1   149
#> [607,]  82   1      0        180        1   193
#> [608,]  87   2      0          5        1   137
#> [609,]  86  12      1          0        1   132
#> [610,]  82  14      1         11        1   103
#> [611,]  86   6      1          0        1   140
#> [612,]  84  16      0         70        1   150
#> [613,]  83  10      1          0        1   190
#> [614,]  86   2      0        180        1   169
#> [615,]  84   3      0          3        1   121
#> [616,]  83  13      1         12        0   170
#> [617,]  82   4      0          4        0   130
#> [618,]  86  13      0        177        0   163
#> [619,]  85   3      0          3        1   113
#> [620,]  86   6      0          6        1   117
#> [621,]  83  20      1          3        1   150
#> [622,]  88   4      0          4        1   115
#> [623,]  85  22      0         22        1   184
#> [624,]  83   9      0         65        1   150
#> [625,]  86   9      1          7        1   142
#> [626,]  87   2      0        180        1   130
#> [627,]  86   6      0         46        0   173
#> [628,]  88   3      0        115        0   110
#> [629,]  83   3      0          3        1   130
#> [630,]  87   8      0          8        1   157
#> [631,]  88   4      0          4        0    86
#> [632,]  89   5      0        119        1   140
#> [633,]  87   6      0        180        1   110
#> [634,]  84   2      0        110        1   174
#> [635,]  87  15      1          9        1   138
#> [636,]  90  14      0         14        1   100
#> [637,]  88   1      0          1        0   135
#> [638,]  86   4      0        180        1   145
#> [639,]  91   8      0          8        0   100
#> [640,]  87   6      1          0        0   125
#> [641,]  88   8      0         50        1   154
#> [642,]  90  11      1         10        1   186
#> [643,]  86  10      0        180        1   137
#> [644,]  90   4      1          0        0   121
#> [645,]  87  43      0        178        1   130
#> [646,]  87   5      0         36        1   150
#> [647,]  88   3      1          2        0   159
#> [648,]  89   3      1          1        1   160
#> [649,]  92   1      0          1        1   167
#> [650,]  87   7      0         74        1   105
#> [651,]  89  12      1          0        1   130
#> [652,]  89  52      0         52        1   130
#> [653,]  91   0      0          0        0     0
#> [654,]  89  14      0        180        1    84
#> [655,]  90  18      0        180        0   188
#> [656,]  91   4      1          0        1   120
#> [657,]  90  19      1         11        1   129
#> [658,]  90   1      0          1        1   118
#> [659,]  93   8      0        179        1   110
#> [660,]  92   4      0         76        1   149
#> [661,]  91   1      0        180        0   158
#> [662,]  90   3      0         67        0   162
#> [663,]  95   8      1          5        1   150
#> [664,]  91  12      0         53        1   212
#> [665,]  93   0      1          0        1   122
#> [666,]  92   5      0         69        0   139
#> [667,]  92   2      0          2        0   112
#> [668,]  93   4      0        180        1   135
#> [669,]  96   3      1          0        1   104
#> [670,]  96  15      1          0        1   140
#> 
#> $y
#>   [1] 180.0+   5.0+ 180.0+   5.0+   2.0+ 180.0+ 180.0+  12.0    5.0+ 180.0+
#>  [11] 180.0+ 180.0+   5.0+   2.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [21] 180.0+ 155.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [31] 180.0+ 180.0+ 150.0  180.0+ 180.0+   6.0+ 180.0+ 180.0+ 180.0+   5.0+
#>  [41] 161.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0+ 180.0+ 180.0+ 177.0+ 180.0+
#>  [51] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  10.0+ 180.0+ 180.0+   7.0  180.0+
#>  [61] 180.0+   2.0    1.0  179.0+ 180.0+ 179.0+ 180.0+ 180.0+   4.0+ 180.0+
#>  [71] 180.0+ 180.0+ 180.0+ 180.0+   7.0   36.0   88.0+ 180.0+   4.0+ 180.0+
#>  [81] 180.0+ 180.0+ 180.0+  77.0  180.0+   5.0  180.0+ 180.0+ 180.0+   4.0+
#>  [91]  85.0  166.0+  99.0  180.0+ 152.0+   7.0+   6.0+ 180.0+ 180.0+  13.0+
#> [101] 171.0+ 180.0+ 180.0+ 174.0+   6.0+   1.0  180.0+   9.0+ 180.0+ 180.0+
#> [111] 180.0+   2.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+  16.0+ 180.0+  15.0+
#> [121] 180.0+  12.0+ 180.0+   8.0    3.0+ 180.0+ 180.0+ 180.0+ 180.0+ 140.0 
#> [131] 180.0+   1.0  165.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+   0.5  180.0+
#> [141] 180.0+ 180.0+ 180.0+ 171.0+  15.0    5.0+   4.0+ 147.0+ 180.0+ 180.0+
#> [151]   5.0+   4.0+   9.0+   1.0    2.0+ 180.0+ 180.0+ 180.0+   2.0  180.0+
#> [161] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   0.5  180.0+ 161.0+ 180.0+ 180.0+
#> [171]   3.0  180.0+ 180.0+  10.0+ 180.0+ 180.0+  45.0   19.0  180.0+ 180.0+
#> [181]  24.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  77.0   13.0+ 180.0+
#> [191] 180.0+ 180.0+ 170.0  180.0+   7.0    7.0+   6.0  180.0+ 180.0+   3.0+
#> [201] 180.0+  30.0  180.0+   5.0   18.0    5.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [211] 180.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+ 180.0+   9.0  180.0+   7.0+
#> [221] 180.0+ 180.0+   1.0  180.0+ 180.0+   4.0+ 167.0  180.0+  17.0  180.0+
#> [231]  12.0  180.0+ 180.0+  14.0+  36.0    3.0+  88.0  180.0+  12.0  180.0+
#> [241] 180.0+ 180.0+ 180.0+  12.0+ 180.0+ 180.0+ 180.0+ 180.0+   9.0  180.0+
#> [251] 180.0+ 180.0+   2.0+  18.0+ 180.0+ 180.0+ 180.0+   3.0+   2.0+ 180.0+
#> [261] 180.0+  13.0  179.0+ 166.0+  14.0+   0.5+ 180.0+ 180.0+ 175.0+ 180.0+
#> [271]   7.0+   8.0    5.0   16.0  180.0+   1.0  180.0+ 180.0+   1.0+  11.0+
#> [281] 180.0+  79.0   80.0   15.0+ 180.0+  15.0  180.0+ 180.0+ 180.0+ 180.0+
#> [291] 180.0+   8.0+ 180.0+ 180.0+ 180.0+ 180.0+   6.0  180.0+  19.0+  12.0 
#> [301]  99.0  179.0+ 180.0+ 180.0+  11.0+ 180.0+   0.5   18.0    7.0+ 180.0+
#> [311] 180.0+ 152.0+ 180.0+ 180.0+  93.0   21.0+ 101.0  180.0+ 180.0+   1.0 
#> [321] 180.0+   0.5  180.0+ 180.0+ 180.0+   7.0+   2.0   45.0    5.0+ 103.0 
#> [331] 180.0+  36.0    5.0+ 180.0+ 180.0+  97.0  180.0+   7.0    8.0+   2.0+
#> [341] 180.0+  18.0  180.0+   7.0  123.0  180.0+  19.0  180.0+   1.0   76.0 
#> [351] 180.0+  10.0+ 180.0+ 180.0+ 162.0    7.0+ 124.0    9.0   12.0  180.0+
#> [361] 180.0+ 152.0   76.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  28.0 
#> [371] 180.0+ 180.0+ 180.0+ 180.0+  16.0+ 180.0+ 180.0+ 180.0+  15.0    3.0+
#> [381]  13.0+ 180.0+ 180.0+ 180.0+   3.0+ 180.0+ 180.0+   0.5  180.0+ 180.0+
#> [391] 180.0+  87.0   12.0  180.0+   4.0+ 180.0+ 180.0+ 180.0+ 175.0    2.0 
#> [401] 180.0+  14.0+  10.0+ 180.0+ 179.0+ 180.0+ 159.0   15.0  180.0+  10.0 
#> [411]   1.0   13.0  180.0+   4.0+ 180.0+  10.0   57.0  180.0+  11.0    3.0+
#> [421]   5.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  34.0  180.0+
#> [431] 177.0+ 180.0+   5.0   11.0+   4.0+ 180.0+   7.0   15.0+  10.0  180.0+
#> [441] 180.0+ 180.0+ 180.0+ 180.0+   6.0  180.0+  85.0   17.0+ 174.0+   4.0 
#> [451]   0.5  180.0+ 180.0+ 180.0+ 180.0+   4.0    1.0   52.0  180.0+ 180.0+
#> [461]  12.0  180.0+   8.0   33.0  180.0+ 180.0+   1.0  180.0+  79.0  168.0+
#> [471] 180.0+ 180.0+  47.0  180.0+  11.0  180.0+   8.0+   7.0   32.0  180.0+
#> [481] 180.0+ 172.0   12.0    8.0  180.0+ 180.0+  80.0  180.0+ 180.0+  11.0 
#> [491] 152.0+   3.0   29.0   24.0   32.0   23.0    6.0    3.0+ 180.0+ 180.0+
#> [501] 180.0+   1.0   11.0    4.0  180.0+   3.0+   2.0+ 180.0+   1.0  180.0+
#> [511]   6.0  138.0  180.0+  71.0    8.0   40.0   59.0   17.0   10.0+ 180.0+
#> [521]  93.0  118.0  173.0  180.0+ 175.0+  22.0   15.0+ 180.0+ 171.0+ 166.0+
#> [531]  20.0+   1.0    3.0+ 180.0+  85.0   10.0    6.0  180.0+   5.0    1.0 
#> [541] 108.0  180.0+ 125.0  180.0+   6.0    9.0+ 180.0+ 180.0+ 180.0+ 177.0+
#> [551]  70.0    4.0  180.0+ 180.0+ 180.0+ 180.0+   7.0+ 180.0+  20.0    8.0+
#> [561]  16.0  180.0+ 180.0+ 180.0+ 177.0+ 180.0+ 180.0+   2.0  128.0  180.0+
#> [571]   4.0    1.0   38.0  180.0+ 180.0+ 180.0+ 180.0+   4.0   71.0    1.0 
#> [581]  19.0   30.0  180.0+ 180.0+ 114.0  180.0+   2.0  180.0+   4.0+ 180.0+
#> [591]   1.0    5.0+   4.0+ 180.0+   3.0   83.0   88.0  126.0    8.0  165.0 
#> [601]   3.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0  180.0+ 174.0 
#> [611]   6.0   70.0  180.0+ 180.0+   3.0   13.0    4.0  177.0    3.0+   6.0+
#> [621]  20.0    4.0   22.0   65.0   11.0  180.0+  46.0  115.0    3.0+   8.0+
#> [631]   4.0  119.0  180.0+ 110.0  180.0+  14.0    1.0+ 180.0+   8.0   25.0 
#> [641]  50.0   11.0  180.0+   4.0  178.0+  36.0   75.0    3.0+   1.0   74.0 
#> [651] 180.0+  52.0    0.5  180.0+ 180.0+   4.0  180.0+   1.0+ 179.0+  76.0 
#> [661] 180.0+  67.0    8.0   53.0    0.5   69.0    2.0  180.0+   3.0   15.0+
#> 
#> $weights
#> NULL
#> 
#> $offset
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Warning: Multiple lambdas have been fit. Lambda will be set to 0.01 (see parameter 's').

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8222156