GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.coxnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.coxnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Caution: This learner is different to learners calling glmnet::cv.glmnet() in that it does not use the internal optimization of parameter lambda. Instead, lambda needs to be tuned by the user (e.g., via mlr3tuning). When lambda is tuned, the glmnet will be trained for each tuning iteration. While fitting the whole path of lambdas would be more efficient, as is done by default in glmnet::glmnet(), tuning/selecting the parameter at prediction time (using parameter s) is currently not supported in mlr3 (at least not in efficient manner). Tuning the s parameter is, therefore, currently discouraged.

When the data are i.i.d. and efficiency is key, we recommend using the respective auto-tuning counterpart in mlr_learners_surv.cv_glmnet(). However, in some situations this is not applicable, usually when data are imbalanced or not i.i.d. (longitudinal, time-series) and tuning requires custom resampling strategies (blocked design, stratification).

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.coxnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exact	logical	FALSE	TRUE, FALSE	-

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	TRUE	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	0.01		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.logistic	character	Newton	Newton, modified.Newton	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvGlmnet

Methods

Public methods

LearnerSurvGlmnet$new()

LearnerSurvGlmnet$selected_features()

LearnerSurvGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.glmnet")
print(learner)
#> 
#> ── <LearnerSurvGlmnet> (surv.glmnet): Regularized Generalized Linear Model ─────
#> • Model: -
#> • Parameters: use_pred_offset=TRUE
#> • Packages: mlr3, mlr3proba, mlr3extralearners, and glmnet
#> • Predict Types: [crank], distr, and lp
#> • Feature Types: logical, integer, and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: offset, selected_features, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#>    Df  %Dev   Lambda
#> 1   0  0.00 0.192100
#> 2   1  0.57 0.175000
#> 3   1  1.06 0.159400
#> 4   3  1.93 0.145300
#> 5   3  5.49 0.132400
#> 6   3  8.04 0.120600
#> 7   3  9.98 0.109900
#> 8   3 11.50 0.100100
#> 9   3 12.72 0.091240
#> 10  3 13.73 0.083130
#> 11  3 14.56 0.075750
#> 12  4 15.27 0.069020
#> 13  4 15.89 0.062890
#> 14  5 16.42 0.057300
#> 15  5 16.89 0.052210
#> 16  5 17.29 0.047570
#> 17  6 17.63 0.043350
#> 18  6 17.95 0.039500
#> 19  6 18.22 0.035990
#> 20  6 18.44 0.032790
#> 21  6 18.64 0.029880
#> 22  6 18.80 0.027220
#> 23  6 18.94 0.024800
#> 24  6 19.06 0.022600
#> 25  6 19.16 0.020590
#> 26  6 19.24 0.018760
#> 27  6 19.31 0.017100
#> 28  6 19.37 0.015580
#> 29  6 19.42 0.014190
#> 30  6 19.46 0.012930
#> 31  6 19.50 0.011780
#> 32  6 19.53 0.010740
#> 33  6 19.55 0.009783
#> 34  6 19.57 0.008914
#> 35  6 19.59 0.008122
#> 36  6 19.61 0.007401
#> 37  6 19.62 0.006743
#> 38  6 19.63 0.006144
#> 39  6 19.64 0.005598
#> 40  6 19.65 0.005101
#> 41  6 19.65 0.004648
#> 42  6 19.66 0.004235
#> 43  6 19.66 0.003859
#> 44  6 19.67 0.003516
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  28   9      0        180        1   107
#>   [2,]  32   5      1          0        1   121
#>   [3,]  35  10      1          9        0   106
#>   [4,]  35   2      0        180        0   121
#>   [5,]  35   2      1          1        1   112
#>   [6,]  37   9      0        180        1   151
#>   [7,]  38  13      1          0        1   161
#>   [8,]  36   1      0        180        1   155
#>   [9,]  38  12      1          8        1   120
#>  [10,]  33   6      1          1        1   115
#>  [11,]  42   3      1          1        1   130
#>  [12,]  37   1      1          0        1   146
#>  [13,]  40   2      1          1        1   148
#>  [14,]  42   2      0        180        1   100
#>  [15,]  40   6      0        180        1   138
#>  [16,]  40  11      1         10        1   120
#>  [17,]  42   2      0        180        0   100
#>  [18,]  43   3      1          0        1   100
#>  [19,]  41   2      1          1        0   166
#>  [20,]  42  15      1         13        1   125
#>  [21,]  40   3      1          1        0   170
#>  [22,]  42  12      1         10        1   170
#>  [23,]  43   2      1          1        1   116
#>  [24,]  44   5      1          1        0   170
#>  [25,]  41  10      1          8        0   150
#>  [26,]  44   3      0        180        0   141
#>  [27,]  41  13      1          1        0   140
#>  [28,]  45   9      1          7        0   110
#>  [29,]  45   6      0        180        1   170
#>  [30,]  43   2      0        180        1   140
#>  [31,]  45   2      0        180        1   140
#>  [32,]  46  15      0        180        0   120
#>  [33,]  47   4      1          3        0   118
#>  [34,]  45   3      0        150        0   130
#>  [35,]  44   3      1          0        1   180
#>  [36,]  46   7      1          2        0   166
#>  [37,]  45   4      1          0        0   124
#>  [38,]  47   6      1          0        1   116
#>  [39,]  46  13      1         10        0   100
#>  [40,]  44   0      1          0        1    96
#>  [41,]  47   4      1          3        1   160
#>  [42,]  43   3      1          0        1   124
#>  [43,]  49   5      0         73        1   136
#>  [44,]  45   5      0          5        0   141
#>  [45,]  46   6      1          0        1   100
#>  [46,]  47   2      0        180        0   108
#>  [47,]  44   9      1          8        1   135
#>  [48,]  45   5      0        180        1   190
#>  [49,]  44   2      0        180        0   142
#>  [50,]  46  15      0        180        1   120
#>  [51,]  45   9      1          0        1   145
#>  [52,]  48   3      0        180        0   154
#>  [53,]  48  12      1         11        0   200
#>  [54,]  47   5      1          3        1   130
#>  [55,]  46   3      1          0        1   119
#>  [56,]  47  10      0         10        1   140
#>  [57,]  50   1      1          0        1   129
#>  [58,]  48   2      1          0        0   184
#>  [59,]  47   7      0        180        0   145
#>  [60,]  50   4      1          1        0   125
#>  [61,]  50   6      1          2        1   140
#>  [62,]  46   3      1          1        1   140
#>  [63,]  46   9      1          9        1   122
#>  [64,]  50   7      0        180        1   110
#>  [65,]  49   2      0          2        0   105
#>  [66,]  51   1      0          1        1   145
#>  [67,]  49  23      0        179        1   112
#>  [68,]  50   7      1          0        1    92
#>  [69,]  50   4      0          4        1   100
#>  [70,]  50   1      1          0        0   150
#>  [71,]  50   9      0        180        0   130
#>  [72,]  49   7      1          4        1    90
#>  [73,]  47   6      0        180        1   162
#>  [74,]  51   8      0        180        1   140
#>  [75,]  50   4      1          1        0   150
#>  [76,]  53   8      0         36        1   160
#>  [77,]  47   2      1          1        0   110
#>  [78,]  49   9      1          3        0   102
#>  [79,]  54  17      1         12        1   102
#>  [80,]  53   5      0         77        0   159
#>  [81,]  53   7      1          0        0   199
#>  [82,]  54   6      1          3        0   129
#>  [83,]  50  14      1         13        0   170
#>  [84,]  53   8      1          7        0   160
#>  [85,]  48   3      1          2        0   150
#>  [86,]  51  25      1          1        0   202
#>  [87,]  49   5      1          2        1   150
#>  [88,]  53   4      0          4        0   140
#>  [89,]  48   6      0        180        0   160
#>  [90,]  48  11      1         10        0   120
#>  [91,]  51  13      0         99        1   160
#>  [92,]  55   6      1          2        1   114
#>  [93,]  55   4      1          2        0   150
#>  [94,]  52   4      0        180        1   180
#>  [95,]  51  13      1         11        0   145
#>  [96,]  50   5      1          4        1   150
#>  [97,]  50   3      0        174        1   153
#>  [98,]  55  28      1         13        1   160
#>  [99,]  49   1      0          1        1   110
#> [100,]  50   7      1          1        0   156
#> [101,]  53   8      1          0        1   130
#> [102,]  56   4      1          1        1   130
#> [103,]  52   5      0        175        1   117
#> [104,]  54   1      0        180        0   162
#> [105,]  54   3      0        180        1   180
#> [106,]  55   6      0        180        0   170
#> [107,]  52  16      0         16        0   152
#> [108,]  52  16      1         14        0   170
#> [109,]  53   4      0        180        1   150
#> [110,]  55   6      1          5        1   138
#> [111,]  55   2      0        134        1   140
#> [112,]  54   3      0        180        0   128
#> [113,]  56   3      0          8        1   139
#> [114,]  55   1      0          2        0   130
#> [115,]  57   3      0          3        0   120
#> [116,]  54   7      1          2        0   129
#> [117,]  54   2      1          1        0   135
#> [118,]  52   9      1          3        0   170
#> [119,]  54   2      1          1        1   176
#> [120,]  57   5      1          3        1   138
#> [121,]  57   1      0          1        1   100
#> [122,]  56   4      1          0        1   140
#> [123,]  55  11      1          7        0   104
#> [124,]  56  14      1         11        0   130
#> [125,]  57  10      0        180        1   170
#> [126,]  58   8      0          8        1   130
#> [127,]  54   5      0        180        1   108
#> [128,]  57   0      0          0        1   150
#> [129,]  53  21      1         13        1   130
#> [130,]  59   3      1          1        0   172
#> [131,]  57   4      0        180        1   119
#> [132,]  58   6      1          0        1    90
#> [133,]  53  15      1         10        1   130
#> [134,]  56   5      0          5        1   150
#> [135,]  54  23      1          8        0   120
#> [136,]  57   4      1          2        1   185
#> [137,]  53   4      0        147        1   145
#> [138,]  53   7      1          0        1   120
#> [139,]  55   5      0          5        1   131
#> [140,]  54   7      1          0        1   141
#> [141,]  56   4      0          4        0   164
#> [142,]  55   5      1          0        0   140
#> [143,]  56   7      1          5        1   120
#> [144,]  55   2      0          2        0   106
#> [145,]  59   9      1          1        1   125
#> [146,]  57   1      0        180        0   148
#> [147,]  59   3      0        180        0   120
#> [148,]  58   4      1          0        1   160
#> [149,]  60   5      1          1        0   138
#> [150,]  55   5      1          0        1   160
#> [151,]  57  10      1          9        0   103
#> [152,]  59   4      1          0        1   152
#> [153,]  61   9      0          9        1   160
#> [154,]  58   4      1          3        0   120
#> [155,]  59   2      1          1        0   140
#> [156,]  61   9      1          8        0   150
#> [157,]  61   3      1          2        1   102
#> [158,]  58   1      0          1        1   100
#> [159,]  61  20      1         13        0   130
#> [160,]  57  13      1         10        0   110
#> [161,]  57   2      1          0        1   116
#> [162,]  58  10      0         10        1   150
#> [163,]  57  11      0        180        1   150
#> [164,]  61   3      0         17        0   143
#> [165,]  56  14      0         45        0   130
#> [166,]  58  19      1         13        1   140
#> [167,]  56  13      1          6        1   158
#> [168,]  56  18      1         11        1   165
#> [169,]  55   4      1          3        1   160
#> [170,]  60  12      1          0        1   114
#> [171,]  55   9      1          7        1   135
#> [172,]  56   8      1          8        0   120
#> [173,]  61  13      1         12        1   130
#> [174,]  59  11      1          8        1   190
#> [175,]  58   5      1          1        1   135
#> [176,]  59  10      0        180        0   160
#> [177,]  58   8      1          5        0   152
#> [178,]  62  10      1          0        1   153
#> [179,]  62   7      1          2        1   180
#> [180,]  61  18      0        170        0   140
#> [181,]  58   8      1          3        1   150
#> [182,]  57   7      0        169        0   180
#> [183,]  61   7      0          7        1   150
#> [184,]  60   7      0          7        0   147
#> [185,]  61   6      0          6        0   134
#> [186,]  57  12      1          9        1   120
#> [187,]  60  17      1          8        1   140
#> [188,]  58   3      1          0        1   146
#> [189,]  62   4      1          3        0   173
#> [190,]  58   2      0         30        0   202
#> [191,]  59   1      0        180        0   155
#> [192,]  59  16      1          9        1   133
#> [193,]  61   5      0          5        1   110
#> [194,]  57  18      1          9        1    93
#> [195,]  58  11      1          9        0   179
#> [196,]  57   2      1          1        0   159
#> [197,]  62  17      1         10        1   180
#> [198,]  58   7      0        180        1   150
#> [199,]  63   3      1          1        0   180
#> [200,]  63   1      0        180        1   130
#> [201,]  63   4      1          3        0   222
#> [202,]  62   3      0        180        1   105
#> [203,]  63   4      0        180        1   190
#> [204,]  63  15      1         10        1   126
#> [205,]  63   4      1          1        0   155
#> [206,]  60  18      1         13        0   132
#> [207,]  59   8      0        180        1   140
#> [208,]  58   9      1          9        0   110
#> [209,]  62   7      0          7        0   150
#> [210,]  59   1      0         22        1   162
#> [211,]  59   4      0        180        0   196
#> [212,]  60   7      1          5        1   141
#> [213,]  59   5      1          1        0   148
#> [214,]  65  13      0        180        1   100
#> [215,]  63   1      0          1        0   130
#> [216,]  62   6      0        180        0   170
#> [217,]  61  15      1         13        0   170
#> [218,]  59   4      0          4        0   149
#> [219,]  60   3      0          3        0   168
#> [220,]  64  10      1          9        0   160
#> [221,]  63  12      1         10        0   200
#> [222,]  59  10      0        180        1   130
#> [223,]  60   8      0         17        1   130
#> [224,]  61   6      1          1        1   117
#> [225,]  64  12      1         11        0   160
#> [226,]  64   6      1          0        1   140
#> [227,]  63  10      1          0        1   148
#> [228,]  63  14      1          9        0   123
#> [229,]  65  36      1         11        0   140
#> [230,]  63   4      1          3        0   162
#> [231,]  66   3      1          1        0   127
#> [232,]  61  10      1          2        1   194
#> [233,]  64  32      1          9        1   160
#> [234,]  63  12      1          9        0   114
#> [235,]  63   7      0        180        0   120
#> [236,]  65   8      1          0        0   168
#> [237,]  64  21      1         10        0   190
#> [238,]  64   9      0        180        0   150
#> [239,]  61   4      0        180        1   113
#> [240,]  63  16      1          7        1   110
#> [241,]  64   7      0        180        1   120
#> [242,]  66   6      1          1        1   130
#> [243,]  63  12      0         12        1   150
#> [244,]  65   3      1          0        1    80
#> [245,]  63   2      1          1        0   180
#> [246,]  62  13      1         11        0   180
#> [247,]  64   2      0          2        0   201
#> [248,]  66  18      1          5        0   142
#> [249,]  66  16      1         11        1   169
#> [250,]  61  14      1          5        0   140
#> [251,]  61  15      1         10        0   130
#> [252,]  63   9      1          8        1   160
#> [253,]  63   2      1          0        0   140
#> [254,]  64  19      1          8        1   160
#> [255,]  65   8      1          0        1   140
#> [256,]  67   6      0        180        1   170
#> [257,]  65  15      1         11        1   160
#> [258,]  64  13      1         12        1   150
#> [259,]  64   6      1          0        1   125
#> [260,]  66   7      1          0        1   115
#> [261,]  64  14      1         13        1   150
#> [262,]  65   3      0          3        0   105
#> [263,]  64   0      0          0        1   148
#> [264,]  67   4      1          3        0   130
#> [265,]  65   2      1          1        1   170
#> [266,]  68   1      0        180        1   166
#> [267,]  64  10      1          9        1   110
#> [268,]  67   8      1          1        1   130
#> [269,]  68   5      0          5        1    90
#> [270,]  66  14      0        180        0   130
#> [271,]  64   1      0          1        1   120
#> [272,]  68  18      0        180        1   260
#> [273,]  65  18      1          3        0   120
#> [274,]  63  10      0         18        1   130
#> [275,]  67  11      0         11        0   150
#> [276,]  68  11      0        180        0   160
#> [277,]  66  12      1         10        1   150
#> [278,]  66  11      1          0        0   100
#> [279,]  65   4      1          2        1   145
#> [280,]  69  12      0         15        1   140
#> [281,]  69  21      1         10        0   180
#> [282,]  69   6      0        180        1   100
#> [283,]  66   9      1          8        0   130
#> [284,]  63   8      0        180        1   120
#> [285,]  68  14      1         13        1   140
#> [286,]  65   8      1          0        1    90
#> [287,]  66   3      0          3        1   138
#> [288,]  69   1      1          0        0   170
#> [289,]  67   1      0        180        1   160
#> [290,]  68  10      1         10        1   150
#> [291,]  65   1      1          0        0   133
#> [292,]  67   7      1          4        1   130
#> [293,]  63   2      1          0        0    99
#> [294,]  67   2      0        180        0   184
#> [295,]  66  19      1         12        1   150
#> [296,]  67  12      1         12        0   160
#> [297,]  69   6      0         99        1   140
#> [298,]  65   4      1          1        0   130
#> [299,]  70  15      1         12        1   132
#> [300,]  64   0      1          0        1   118
#> [301,]  69   4      1          3        1   150
#> [302,]  65  13      1         12        1   130
#> [303,]  69  17      1         10        0   140
#> [304,]  69   8      0         93        0   140
#> [305,]  64  21      0         21        1   155
#> [306,]  65   1      0          1        1   120
#> [307,]  65   6      0        101        1   115
#> [308,]  68   4      0          4        1   190
#> [309,]  71   3      0          5        0   112
#> [310,]  70   7      1          0        1   190
#> [311,]  67   2      0        180        0   128
#> [312,]  66   9      1          3        1   151
#> [313,]  66   1      1          1        1   165
#> [314,]  70   4      1          0        1   180
#> [315,]  66   4      0        180        0   130
#> [316,]  67  10      1          9        0   200
#> [317,]  68  18      1         14        1   170
#> [318,]  69   0      0          0        1   148
#> [319,]  65   2      0        180        0   130
#> [320,]  68   7      1          0        1   150
#> [321,]  69   3      1          2        0   151
#> [322,]  65  14      1         13        1   150
#> [323,]  69   8      0        180        1   180
#> [324,]  71   7      0          7        0   230
#> [325,]  66   2      0          2        1   228
#> [326,]  71   6      0         45        1   158
#> [327,]  71   3      0        103        0   133
#> [328,]  69   3      0          3        1   130
#> [329,]  70  22      1         13        0   103
#> [330,]  67   1      0         36        1   104
#> [331,]  67   5      0          5        0   130
#> [332,]  68   6      0        180        0   145
#> [333,]  69   8      1          5        1   195
#> [334,]  69   6      1          4        1   174
#> [335,]  72   3      1          0        1   132
#> [336,]  72   7      0          7        1   110
#> [337,]  67   3      0        180        0   110
#> [338,]  69  19      0        180        0   130
#> [339,]  68  18      0         18        1   100
#> [340,]  66   2      0        180        0   130
#> [341,]  67   7      1          4        0   122
#> [342,]  68   2      0          7        1   130
#> [343,]  69   8      1          2        0   121
#> [344,]  67  13      1          9        0   130
#> [345,]  70   3      0        123        0   130
#> [346,]  72   5      1          4        0   170
#> [347,]  67  22      1          1        1   140
#> [348,]  68   3      0         19        0   135
#> [349,]  67  12      1          8        0   120
#> [350,]  69   1      0          1        1   110
#> [351,]  67   1      0          1        1    60
#> [352,]  67   4      0         60        1   136
#> [353,]  69   5      0         76        0   120
#> [354,]  72  13      1         11        1   195
#> [355,]  68  10      1          8        1   160
#> [356,]  72  30      1          0        1   145
#> [357,]  70   7      0          7        0   102
#> [358,]  71   6      0          9        0   120
#> [359,]  70  11      0        180        1   210
#> [360,]  72  19      1          8        0   120
#> [361,]  67   8      0        180        1   170
#> [362,]  67   5      1          0        1   147
#> [363,]  67   9      0        180        0   158
#> [364,]  73  13      0        152        1   130
#> [365,]  70   5      0        180        0   150
#> [366,]  72   2      0          2        1   100
#> [367,]  67   4      1          1        0   134
#> [368,]  72   6      1          5        0   115
#> [369,]  71   1      0        173        1   188
#> [370,]  68  23      0        180        1   220
#> [371,]  70   3      0        180        0   121
#> [372,]  68   4      1          3        0   210
#> [373,]  72   5      0         28        0   120
#> [374,]  71   5      0        180        0   191
#> [375,]  73   6      0        180        1   117
#> [376,]  69  16      1         10        1   140
#> [377,]  69   8      1          1        0   164
#> [378,]  68   7      0        180        1   130
#> [379,]  72  16      1          1        1   130
#> [380,]  70   4      0        180        0   180
#> [381,]  69   1      1          0        0   155
#> [382,]  73   6      1          0        1   270
#> [383,]  72   8      1          1        1   150
#> [384,]  71   2      1          0        1   180
#> [385,]  73   7      0          7        1   140
#> [386,]  68  15      1         13        1   130
#> [387,]  70  13      1          9        0   100
#> [388,]  72   6      0        180        1   130
#> [389,]  74   8      1          0        1    85
#> [390,]  73   4      0        180        1   154
#> [391,]  71   3      1          1        0   150
#> [392,]  71  15      1         11        0   165
#> [393,]  68   9      0        180        1   120
#> [394,]  71  20      1         10        0   140
#> [395,]  73   3      1          0        1   136
#> [396,]  70   5      1          0        1   190
#> [397,]  71  17      1         11        0   160
#> [398,]  69  12      1          1        1   149
#> [399,]  70  26      1         11        1   120
#> [400,]  73   4      0         58        1   160
#> [401,]  72   5      1          3        1   160
#> [402,]  70   3      0        180        1   154
#> [403,]  73   6      0        180        0   110
#> [404,]  71   7      1          2        0   143
#> [405,]  72   8      1          0        1   140
#> [406,]  73  17      1         11        0   140
#> [407,]  70   4      1          0        1   140
#> [408,]  74   7      1          0        1   117
#> [409,]  69   7      0        180        1   144
#> [410,]  72  15      1         13        0   156
#> [411,]  70   8      0          8        0   120
#> [412,]  75   1      0          1        0   133
#> [413,]  75   2      1          1        0   145
#> [414,]  72  10      1          9        1   160
#> [415,]  73  10      1         10        1   120
#> [416,]  74  15      1          9        1   179
#> [417,]  71   2      0         10        1   112
#> [418,]  73   1      0          1        1    80
#> [419,]  75   9      1          7        0   140
#> [420,]  75  13      1          1        1   130
#> [421,]  71   4      0          4        0   134
#> [422,]  72  15      1         12        1   120
#> [423,]  72   7      0         57        1   145
#> [424,]  73  10      0        180        0   162
#> [425,]  72  11      0         11        1   140
#> [426,]  70   3      0          3        0   150
#> [427,]  76  25      1         12        1   170
#> [428,]  73  12      1         12        1   140
#> [429,]  72   2      0        180        0   120
#> [430,]  75   1      0        180        1   140
#> [431,]  71   3      1          0        0   144
#> [432,]  73   5      0        180        0   126
#> [433,]  73   4      0        180        0   124
#> [434,]  74  34      1          8        1   233
#> [435,]  76   3      1          0        1   120
#> [436,]  72   3      0        180        0   160
#> [437,]  76   5      0          5        1   130
#> [438,]  77  11      0         11        1   150
#> [439,]  77   4      0          4        0   185
#> [440,]  75   3      1          1        0   180
#> [441,]  72   7      1          2        0   142
#> [442,]  73  15      0         15        1   160
#> [443,]  73  10      1         10        0   124
#> [444,]  74   7      0        180        1   150
#> [445,]  76   1      0        180        0   114
#> [446,]  76   8      1          0        1   141
#> [447,]  73   6      0          6        1   114
#> [448,]  72   4      0         85        1   120
#> [449,]  76  17      1          0        1   200
#> [450,]  73   4      1          3        1   125
#> [451,]  76  13      1         10        0   110
#> [452,]  75   4      1          0        1   122
#> [453,]  75   7      0          7        0   190
#> [454,]  75   0      0          0        1   130
#> [455,]  73  13      1         11        0   195
#> [456,]  76  13      1          8        1   148
#> [457,]  74   6      0        180        0   160
#> [458,]  76   4      0          4        1   155
#> [459,]  75   1      0          1        1   125
#> [460,]  74   2      0        180        0   111
#> [461,]  78  12      1         11        1   160
#> [462,]  76  44      1         10        0   105
#> [463,]  76   5      0        180        0   185
#> [464,]  74  10      1          0        1   135
#> [465,]  74   8      1          8        1   170
#> [466,]  75   9      0        180        1   140
#> [467,]  73  10      1          9        0   146
#> [468,]  77  12      0        180        0   130
#> [469,]  78   5      1          0        1   170
#> [470,]  73   7      1          0        0   174
#> [471,]  74   9      1          8        0   200
#> [472,]  75   6      0        180        0   150
#> [473,]  74   2      1          0        1   130
#> [474,]  78  18      0         18        1   144
#> [475,]  77   3      0        180        0   110
#> [476,]  73   8      1          1        1   162
#> [477,]  74   2      0        180        0   100
#> [478,]  78   8      1          6        1   110
#> [479,]  78  32      1          9        1   198
#> [480,]  79   6      0        180        0   170
#> [481,]  80  10      1          6        1   147
#> [482,]  78   0      0        180        1   212
#> [483,]  78  13      1          5        0   130
#> [484,]  75   5      0        119        1   150
#> [485,]  75  12      1          1        1   120
#> [486,]  78  15      0        180        1   270
#> [487,]  80   8      0          8        1   120
#> [488,]  75  13      1          6        0   150
#> [489,]  74  10      1          8        0   135
#> [490,]  76   1      0          1        1    83
#> [491,]  78   2      1          1        0   130
#> [492,]  75   4      1          0        0   212
#> [493,]  77   2      1          0        1   143
#> [494,]  78  10      0        180        1   130
#> [495,]  77  24      0         24        1   160
#> [496,]  79   8      0         32        1   120
#> [497,]  80   6      0          6        1   150
#> [498,]  78   6      1          0        1   240
#> [499,]  79  11      0        180        0   160
#> [500,]  79   2      1          0        1   121
#> [501,]  78  14      1          0        1   140
#> [502,]  81   1      0          1        0   130
#> [503,]  78  11      1          8        1   118
#> [504,]  79   4      0          4        1   125
#> [505,]  76  12      1         10        1   127
#> [506,]  80   3      1          0        1   120
#> [507,]  75   2      1          1        1   204
#> [508,]  78  11      0        180        1   135
#> [509,]  76   1      0          1        1   140
#> [510,]  77  31      1          3        1   161
#> [511,]  76   1      0          1        1    90
#> [512,]  79   3      0          3        0   120
#> [513,]  77   7      0        180        1   170
#> [514,]  77   6      0          6        1   144
#> [515,]  79   4      1          0        1   120
#> [516,]  80  15      1         12        1   150
#> [517,]  77   9      1          4        0   141
#> [518,]  82   5      0          8        1   120
#> [519,]  80  40      1          0        1   138
#> [520,]  76   7      0        161        0   151
#> [521,]  79  10      0         10        1   120
#> [522,]  80  15      1          0        1    90
#> [523,]  81   4      1          2        1   126
#> [524,]  80   6      0        173        1   160
#> [525,]  79   1      0         37        1   140
#> [526,]  81   3      0        180        0   184
#> [527,]  78   7      0          7        1   147
#> [528,]  77  13      1          0        1   190
#> [529,]  80   5      1          1        1   108
#> [530,]  79   3      0          3        1   101
#> [531,]  78  26      1          5        0   194
#> [532,]  78   3      1          1        1   152
#> [533,]  77  10      1          8        1   130
#> [534,]  77   5      0         85        0   188
#> [535,]  80   2      1          1        0   168
#> [536,]  78   2      0        180        0   148
#> [537,]  80   5      0          5        1   130
#> [538,]  82   1      1          0        1    82
#> [539,]  77   4      0        180        1    98
#> [540,]  82  21      1          2        0   155
#> [541,]  84  22      1         10        0   180
#> [542,]  79   4      0          4        1   121
#> [543,]  80   6      0          6        1   110
#> [544,]  82   5      0        180        0   110
#> [545,]  83   5      0        180        0   148
#> [546,]  79   7      1          6        0   130
#> [547,]  83   4      0        103        0    97
#> [548,]  81  11      1          8        0   160
#> [549,]  81   5      0        177        0    41
#> [550,]  80  11      1          8        0   170
#> [551,]  79   4      0          4        1   183
#> [552,]  82   8      1          1        0   128
#> [553,]  81  15      0        180        1   140
#> [554,]  80   7      1          0        1   146
#> [555,]  84   5      1          1        1    85
#> [556,]  83   8      0          8        0   115
#> [557,]  81  16      0         16        1   110
#> [558,]  80   6      1          0        1   150
#> [559,]  80  11      1          8        0   110
#> [560,]  85   4      0        180        0    90
#> [561,]  81   2      1          1        0   198
#> [562,]  83   2      0          2        1   155
#> [563,]  82   6      0        128        1   100
#> [564,]  84   4      0        167        0   198
#> [565,]  84   5      0        180        1   203
#> [566,]  84   1      0         38        1   205
#> [567,]  81   4      0         90        1   138
#> [568,]  79   9      1          8        0   150
#> [569,]  85   3      1          2        1   160
#> [570,]  84   4      0         89        1   129
#> [571,]  80   2      1          0        1   130
#> [572,]  79   4      0          4        1    60
#> [573,]  80   6      0         71        1   189
#> [574,]  83   1      0          1        1   100
#> [575,]  79  14      1          0        0   110
#> [576,]  83   3      0        114        0    98
#> [577,]  81  14      1         12        1   128
#> [578,]  83   2      0        154        0   130
#> [579,]  82   0      0          2        1   100
#> [580,]  85   9      1          6        1   160
#> [581,]  81   4      0          4        0   175
#> [582,]  82  16      1          8        0   103
#> [583,]  81   4      0          4        0   160
#> [584,]  86  12      0        180        1   120
#> [585,]  83  12      1          2        1   170
#> [586,]  81  19      1         14        0   120
#> [587,]  82   3      1          2        0   130
#> [588,]  82  15      1          0        0   183
#> [589,]  80   2      0         88        0   135
#> [590,]  83   7      0        126        0   135
#> [591,]  86   8      0          8        1   132
#> [592,]  81  16      1          9        0   180
#> [593,]  84   6      0        165        0   145
#> [594,]  82   9      0        180        1   134
#> [595,]  84   3      0        180        1   120
#> [596,]  83   9      0        180        1   149
#> [597,]  83   4      0          4        0   130
#> [598,]  87   2      0          5        1   137
#> [599,]  86  12      1          0        1   132
#> [600,]  86   6      1          0        1   140
#> [601,]  84  16      0         70        1   150
#> [602,]  83  19      0         43        0   150
#> [603,]  84   3      1          2        0   125
#> [604,]  83  10      1          0        1   190
#> [605,]  84   3      0          3        1   121
#> [606,]  83  13      1         12        0   170
#> [607,]  84   7      1          2        0   148
#> [608,]  87   2      0        180        0   113
#> [609,]  86   4      0         38        1   122
#> [610,]  82   4      0          4        0   130
#> [611,]  86  13      0        177        0   163
#> [612,]  86   6      0          6        1   117
#> [613,]  86   6      1          1        0   112
#> [614,]  88   4      0          4        0   100
#> [615,]  85  22      0         22        1   184
#> [616,]  87   2      0        180        1   130
#> [617,]  86   6      0         46        0   173
#> [618,]  88   3      0        115        0   110
#> [619,]  88   2      0        180        1    68
#> [620,]  87   8      0          8        1   157
#> [621,]  88   4      0          4        0    86
#> [622,]  89   4      0          4        1   153
#> [623,]  89   5      0        119        1   140
#> [624,]  87   1      0          1        0   170
#> [625,]  84   8      0        180        1   119
#> [626,]  85   8      0          8        1   136
#> [627,]  84   2      0        110        1   174
#> [628,]  87  29      0         29        1    97
#> [629,]  87  15      1          9        1   138
#> [630,]  84   0      0        180        1   136
#> [631,]  89  10      0         46        1   170
#> [632,]  87   2      0        180        0   160
#> [633,]  87   6      1          0        0   125
#> [634,]  91  10      0        145        0   135
#> [635,]  86   3      1          0        1    80
#> [636,]  88   7      0         24        0   119
#> [637,]  90  11      1         10        1   186
#> [638,]  86   9      1          7        0   130
#> [639,]  91   1      0          1        1    74
#> [640,]  87  43      0        178        1   130
#> [641,]  90   5      1          0        1   125
#> [642,]  88   3      1          2        0   159
#> [643,]  92   1      0          1        1   167
#> [644,]  91   3      0         33        1   137
#> [645,]  88   5      0        158        0   100
#> [646,]  87   7      0         74        1   105
#> [647,]  89  12      1          0        1   130
#> [648,]  89   2      0        168        0   118
#> [649,]  91   5      0        169        1   176
#> [650,]  89  52      0         52        1   130
#> [651,]  92   7      0          7        1   110
#> [652,]  89   4      0          4        1   159
#> [653,]  91   0      0          0        0     0
#> [654,]  89  14      0        180        1    84
#> [655,]  90  18      0        180        0   188
#> [656,]  91   4      1          0        1   120
#> [657,]  94   6      0         50        0    78
#> [658,]  90   1      0          1        1   118
#> [659,]  91   1      0        180        0   158
#> [660,]  90  16      0         16        1   106
#> [661,]  90   3      0         67        0   162
#> [662,]  95   8      1          5        1   150
#> [663,]  94   3      0         26        1   144
#> [664,]  91  12      0         53        1   212
#> [665,]  91   7      0          7        0   135
#> [666,]  93   0      1          0        1   122
#> [667,]  92   5      0         69        0   139
#> [668,]  92   2      0          2        0   112
#> [669,]  93   4      0        180        1   135
#> [670,]  96   3      1          0        1   104
#> 
#> $y
#>   [1] 180.0+   5.0+ 180.0+ 180.0+   2.0+ 180.0+ 180.0+ 180.0+  12.0  180.0+
#>  [11] 180.0+ 180.0+   2.0+ 180.0+ 180.0+ 180.0+ 180.0+   3.0  180.0+ 180.0+
#>  [21] 180.0+ 180.0+   2.0+ 155.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [31] 180.0+ 180.0+ 180.0+ 150.0  180.0+ 180.0+ 180.0+   6.0+ 180.0+ 180.0+
#>  [41] 180.0+ 180.0+  73.0    5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [51] 177.0+ 180.0+ 180.0+ 180.0+ 180.0+  10.0+ 172.0+ 180.0+ 180.0+ 180.0+
#>  [61] 180.0+ 180.0+ 180.0+ 180.0+   2.0    1.0  179.0+ 180.0+   4.0+ 180.0+
#>  [71] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  36.0  180.0+ 180.0+ 180.0+  77.0 
#>  [81] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+ 180.0+
#>  [91]  99.0    6.0+ 180.0+ 180.0+  13.0+ 171.0+ 174.0+  28.0    1.0  180.0+
#> [101] 180.0+ 180.0+ 175.0+ 180.0+ 180.0+ 180.0+  16.0+  16.0  180.0+ 180.0+
#> [111] 134.0+ 180.0+   8.0    2.0    3.0+ 180.0+ 180.0+ 180.0+ 180.0+ 140.0 
#> [121]   1.0  165.0  180.0+ 180.0+ 180.0+   8.0+ 180.0+   0.5  180.0+ 180.0+
#> [131] 180.0+ 180.0+ 180.0+   5.0+ 180.0+   4.0+ 147.0+ 180.0+   5.0+ 180.0+
#> [141]   4.0+ 180.0+ 180.0+   2.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [151] 180.0+ 180.0+   9.0+ 180.0+ 180.0+ 180.0+   3.0    1.0  180.0+ 180.0+
#> [161] 180.0+  10.0+ 180.0+  17.0   45.0   19.0  180.0+ 180.0+ 180.0+ 172.0+
#> [171]  24.0    8.0  180.0+ 180.0+ 180.0+ 180.0+   8.0+ 180.0+ 180.0+ 170.0 
#> [181] 180.0+ 169.0    7.0    7.0+   6.0  180.0+ 180.0+   3.0+ 180.0+  30.0 
#> [191] 180.0+ 180.0+   5.0   18.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [201] 180.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+ 180.0+   9.0    7.0+  22.0 
#> [211] 180.0+  84.0  180.0+ 180.0+   1.0  180.0+ 180.0+   4.0+   3.0+ 167.0 
#> [221] 180.0+ 180.0+  17.0  180.0+  12.0  180.0+ 180.0+  14.0+  36.0  180.0+
#> [231]   3.0+  88.0  180.0+  12.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [241] 180.0+ 180.0+  12.0    3.0  180.0+ 180.0+   2.0+  18.0+ 180.0+ 180.0+
#> [251] 180.0+ 180.0+   2.0+ 103.0   15.0  180.0+ 180.0+  13.0  180.0+ 179.0+
#> [261]  14.0+   3.0    0.5+ 180.0+ 175.0+ 180.0+ 180.0+   8.0    5.0  180.0+
#> [271]   1.0  180.0+ 123.0+  18.0   11.0+ 180.0+  80.0  180.0+   4.0+  15.0 
#> [281] 174.0+ 180.0+ 180.0+ 180.0+ 180.0+   8.0+   3.0  175.0  180.0+  10.0 
#> [291] 180.0+ 180.0+ 180.0+ 180.0+  19.0+  12.0   99.0  180.0+ 180.0+   0.5 
#> [301] 152.0+ 180.0+ 180.0+  93.0   21.0+   1.0  101.0    4.0    5.0    7.0+
#> [311] 180.0+ 180.0+   1.0  180.0+ 180.0+ 174.0+ 180.0+   0.5  180.0+ 180.0+
#> [321] 180.0+  14.0+ 180.0+   7.0+   2.0   45.0  103.0    3.0+ 180.0+  36.0 
#> [331]   5.0+ 180.0+ 180.0+  97.0  180.0+   7.0  180.0+ 180.0+  18.0  180.0+
#> [341]   7.0    7.0    8.0+  13.0+ 123.0  180.0+  51.0   19.0  180.0+   1.0 
#> [351]   1.0   60.0   76.0  132.0   10.0+ 162.0    7.0+   9.0  180.0+ 180.0+
#> [361] 180.0+ 180.0+ 180.0+ 152.0  180.0+   2.0   76.0  180.0+ 173.0+ 180.0+
#> [371] 180.0+ 180.0+  28.0  180.0+ 180.0+  16.0+ 180.0+ 180.0+  16.0+ 180.0+
#> [381] 180.0+   6.0  180.0+ 180.0+   7.0+  15.0   13.0+ 180.0+ 180.0+ 180.0+
#> [391]   3.0+ 180.0+ 180.0+  20.0  180.0+ 180.0+ 180.0+  12.0  180.0+  58.0 
#> [401] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [411]   8.0+   1.0  180.0+ 159.0   15.0  180.0+  10.0    1.0  180.0+  13.0 
#> [421]   4.0+ 180.0+  57.0  180.0+  11.0    3.0+ 180.0+  12.0  180.0+ 180.0+
#> [431] 180.0+ 180.0+ 180.0+  34.0  180.0+ 180.0+   5.0   11.0+   4.0+ 180.0+
#> [441]   7.0   15.0+  10.0  180.0+ 180.0+ 180.0+   6.0   85.0   17.0+ 180.0+
#> [451] 174.0+   4.0    7.0    0.5  180.0+ 180.0+ 180.0+   4.0    1.0  180.0+
#> [461]  12.0  180.0+ 180.0+ 180.0+   8.0  180.0+ 180.0+ 180.0+ 180.0+   7.0+
#> [471] 168.0+ 180.0+ 176.0+  18.0  180.0+ 180.0+ 180.0+   8.0+  32.0  180.0+
#> [481]  10.0  180.0+ 172.0  119.0   12.0  180.0+   8.0  180.0+ 180.0+   1.0 
#> [491] 180.0+   4.0+   2.0  180.0+  24.0   32.0    6.0  180.0+ 180.0+ 180.0+
#> [501] 180.0+   1.0   11.0    4.0  180.0+   3.0+   2.0+ 180.0+   1.0  171.0 
#> [511]   1.0    3.0  180.0+   6.0  138.0  180.0+  71.0    8.0   40.0  161.0 
#> [521]  10.0+ 180.0+  93.0  173.0   37.0  180.0+   7.0+  22.0    5.0+   3.0 
#> [531] 171.0+   3.0+  10.0   85.0   10.0  180.0+   5.0    1.0  180.0+ 180.0+
#> [541] 180.0+   4.0    6.0  180.0+ 180.0+ 180.0+ 103.0  180.0+ 177.0+ 169.0 
#> [551]   4.0  180.0+ 180.0+   7.0+ 180.0+   8.0+  16.0  180.0+ 180.0+ 180.0+
#> [561] 180.0+   2.0  128.0  167.0  180.0+  38.0   90.0  180.0+ 180.0+  89.0 
#> [571] 180.0+   4.0   71.0    1.0  180.0+ 114.0  180.0+ 154.0    2.0  180.0+
#> [581]   4.0+  16.0+   4.0+ 180.0+  77.0  180.0+   3.0   83.0   88.0  126.0 
#> [591]   8.0  180.0+ 165.0  180.0+ 180.0+ 180.0+   4.0+   5.0  180.0+   6.0 
#> [601]  70.0   43.0  180.0+ 180.0+   3.0   13.0  180.0+ 180.0+  38.0    4.0 
#> [611] 177.0    6.0+   6.0+   4.0+  22.0  180.0+  46.0  115.0  180.0+   8.0+
#> [621]   4.0    4.0  119.0    1.0+ 180.0+   8.0  110.0   29.0  180.0+ 180.0+
#> [631]  46.0  180.0+  25.0  145.0    3.0   24.0   11.0  180.0+   1.0  178.0+
#> [641]  89.0   75.0    1.0   33.0  158.0   74.0  180.0+ 168.0  169.0   52.0 
#> [651]   7.0    4.0    0.5  180.0+ 180.0+   4.0   50.0    1.0+ 180.0+  16.0 
#> [661]  67.0    8.0   26.0   53.0    7.0+   0.5   69.0    2.0  180.0+   3.0 
#> 
#> $weights
#> NULL
#> 
#> $offset
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Warning: Multiple lambdas have been fit. Lambda will be set to 0.01 (see parameter 's').

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8311192