GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.coxnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.coxnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Caution: This learner is different to learners calling glmnet::cv.glmnet() in that it does not use the internal optimization of parameter lambda. Instead, lambda needs to be tuned by the user (e.g., via mlr3tuning). When lambda is tuned, the glmnet will be trained for each tuning iteration. While fitting the whole path of lambdas would be more efficient, as is done by default in glmnet::glmnet(), tuning/selecting the parameter at prediction time (using parameter s) is currently not supported in mlr3 (at least not in efficient manner). Tuning the s parameter is, therefore, currently discouraged.

When the data are i.i.d. and efficiency is key, we recommend using the respective auto-tuning counterpart in mlr_learners_surv.cv_glmnet(). However, in some situations this is not applicable, usually when data are imbalanced or not i.i.d. (longitudinal, time-series) and tuning requires custom resampling strategies (blocked design, stratification).

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.coxnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exact	logical	FALSE	TRUE, FALSE	-

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	TRUE	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	0.01		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.logistic	character	Newton	Newton, modified.Newton	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvGlmnet

Methods

Public methods

LearnerSurvGlmnet$new()

LearnerSurvGlmnet$selected_features()

LearnerSurvGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.glmnet")
print(learner)
#> 
#> ── <LearnerSurvGlmnet> (surv.glmnet): Regularized Generalized Linear Model ─────
#> • Model: -
#> • Parameters: use_pred_offset=TRUE
#> • Packages: mlr3, mlr3proba, mlr3extralearners, and glmnet
#> • Predict Types: [crank], distr, and lp
#> • Feature Types: logical, integer, and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: offset, selected_features, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#>    Df  %Dev   Lambda
#> 1   0  0.00 0.200600
#> 2   1  0.62 0.182800
#> 3   1  1.14 0.166500
#> 4   2  1.77 0.151700
#> 5   2  2.50 0.138300
#> 6   3  4.77 0.126000
#> 7   3  7.21 0.114800
#> 8   3  9.05 0.104600
#> 9   3 10.48 0.095300
#> 10  3 11.62 0.086830
#> 11  3 12.56 0.079120
#> 12  3 13.33 0.072090
#> 13  4 13.98 0.065690
#> 14  4 14.54 0.059850
#> 15  4 15.01 0.054530
#> 16  4 15.40 0.049690
#> 17  4 15.73 0.045280
#> 18  4 16.01 0.041250
#> 19  5 16.26 0.037590
#> 20  5 16.49 0.034250
#> 21  5 16.68 0.031210
#> 22  5 16.84 0.028430
#> 23  5 16.98 0.025910
#> 24  6 17.10 0.023610
#> 25  6 17.20 0.021510
#> 26  6 17.29 0.019600
#> 27  6 17.36 0.017860
#> 28  6 17.43 0.016270
#> 29  6 17.48 0.014830
#> 30  6 17.52 0.013510
#> 31  6 17.56 0.012310
#> 32  6 17.59 0.011220
#> 33  6 17.61 0.010220
#> 34  6 17.64 0.009311
#> 35  6 17.65 0.008484
#> 36  6 17.67 0.007730
#> 37  6 17.68 0.007043
#> 38  6 17.69 0.006418
#> 39  6 17.70 0.005848
#> 40  6 17.71 0.005328
#> 41  6 17.71 0.004855
#> 42  6 17.72 0.004423
#> 43  6 17.72 0.004031
#> 44  6 17.73 0.003672
#> 45  6 17.73 0.003346
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  28   9      0        180        1   107
#>   [2,]  33   2      0          2        0   150
#>   [3,]  34   5      0          5        0   120
#>   [4,]  35   2      0        180        0   121
#>   [5,]  37   9      0        180        1   151
#>   [6,]  38  13      1          0        1   161
#>   [7,]  38   2      0        115        0   150
#>   [8,]  36   1      0        180        1   155
#>   [9,]  35   0      0        180        1   119
#>  [10,]  38  12      1          8        1   120
#>  [11,]  33   6      1          1        1   115
#>  [12,]  38  16      1         10        0   160
#>  [13,]  38  12      1         11        1    92
#>  [14,]  40  12      1          9        0   153
#>  [15,]  42   3      1          1        1   130
#>  [16,]  37   1      1          0        1   146
#>  [17,]  40   2      1          1        1   148
#>  [18,]  42   2      0        180        1   100
#>  [19,]  38   5      1          3        0   125
#>  [20,]  42   2      0          2        0   140
#>  [21,]  40   6      0        180        1   138
#>  [22,]  40  11      1         10        1   120
#>  [23,]  40   1      1          0        1   145
#>  [24,]  43   4      1          0        1   130
#>  [25,]  42  15      1         13        1   125
#>  [26,]  40   3      1          1        0   170
#>  [27,]  42  12      1         10        1   170
#>  [28,]  43   2      1          1        1   116
#>  [29,]  42   2      0        180        1   124
#>  [30,]  44   5      1          1        0   170
#>  [31,]  45   3      0        180        1   154
#>  [32,]  41  10      1          8        0   150
#>  [33,]  44   3      0        180        0   141
#>  [34,]  41   5      1          4        1   141
#>  [35,]  44   2      1          1        1   150
#>  [36,]  45   2      0        180        1   140
#>  [37,]  46   2      1          1        0   126
#>  [38,]  47   4      1          3        0   118
#>  [39,]  48  15      0        180        1   160
#>  [40,]  45   3      0        150        0   130
#>  [41,]  44   3      1          0        1   180
#>  [42,]  46   7      1          2        0   166
#>  [43,]  43  29      0        180        1   180
#>  [44,]  45   4      1          0        0   124
#>  [45,]  47   6      1          0        1   116
#>  [46,]  46  13      1         10        0   100
#>  [47,]  47   4      1          3        1   160
#>  [48,]  43   3      1          0        1   124
#>  [49,]  45   8      1          0        1   117
#>  [50,]  49   5      0         73        1   136
#>  [51,]  46   6      1          0        1   100
#>  [52,]  44   4      1          0        1   114
#>  [53,]  47   2      0        180        0   108
#>  [54,]  44   9      1          8        1   135
#>  [55,]  45   5      0        180        1   190
#>  [56,]  46   4      0        180        1   121
#>  [57,]  44   2      0        180        0   142
#>  [58,]  46  15      0        180        1   120
#>  [59,]  48   3      0        180        0   154
#>  [60,]  48  12      1         11        0   200
#>  [61,]  47   5      1          3        1   130
#>  [62,]  47   9      1          6        0   170
#>  [63,]  46   3      1          0        1   119
#>  [64,]  49   4      0        180        0   117
#>  [65,]  47  10      0         10        1   140
#>  [66,]  48   2      1          0        0   184
#>  [67,]  47   7      0        180        0   145
#>  [68,]  50   4      1          1        0   125
#>  [69,]  50   6      1          2        1   140
#>  [70,]  49   7      1          7        1   110
#>  [71,]  50   7      0        180        1   110
#>  [72,]  51   1      0          1        1   145
#>  [73,]  49  15      1         11        1   160
#>  [74,]  47   2      0        180        0   150
#>  [75,]  49  23      0        179        1   112
#>  [76,]  46   6      1          0        1   156
#>  [77,]  50   7      1          0        1    92
#>  [78,]  50   9      0        180        0   130
#>  [79,]  47   8      0        180        0   160
#>  [80,]  51   8      0        180        1   140
#>  [81,]  46   3      0        180        1   120
#>  [82,]  50   4      1          1        0   150
#>  [83,]  48   7      1          0        1   110
#>  [84,]  48  17      1         10        0   111
#>  [85,]  47   2      1          1        0   110
#>  [86,]  49   9      1          3        0   102
#>  [87,]  53   5      0        180        1   140
#>  [88,]  53   7      1          0        0   199
#>  [89,]  50   2      0          5        1   106
#>  [90,]  50  14      1         13        0   170
#>  [91,]  53   8      1          7        0   160
#>  [92,]  48   3      1          2        0   150
#>  [93,]  51  25      1          1        0   202
#>  [94,]  48  11      1         10        0   120
#>  [95,]  53   4      1          0        1   156
#>  [96,]  51  13      0         99        1   160
#>  [97,]  49  16      0         16        0   125
#>  [98,]  55   3      1          1        0   150
#>  [99,]  54  23      1         10        0   131
#> [100,]  52   7      1          2        0   154
#> [101,]  55   6      1          2        1   114
#> [102,]  55   4      1          2        0   150
#> [103,]  51  13      1         11        0   145
#> [104,]  50   5      1          4        1   150
#> [105,]  54   4      1          0        1   121
#> [106,]  52   4      0        180        0   183
#> [107,]  50   3      0        174        1   153
#> [108,]  55  28      1         13        1   160
#> [109,]  49   6      1          0        1   130
#> [110,]  49   1      0          1        1   110
#> [111,]  50   7      1          1        0   156
#> [112,]  53   9      0          9        1    95
#> [113,]  53   8      1          0        1   130
#> [114,]  50   7      1          0        1   127
#> [115,]  56   4      1          1        1   130
#> [116,]  52   5      0        175        1   117
#> [117,]  55   1      0        180        0   127
#> [118,]  54   1      0        180        0   162
#> [119,]  54   7      1          0        1   100
#> [120,]  55   5      1          4        1   120
#> [121,]  52   8      0        180        0   119
#> [122,]  53  18      1          9        1   150
#> [123,]  54   3      0        180        1   180
#> [124,]  52  16      1         14        0   170
#> [125,]  53   4      0        180        1   150
#> [126,]  55   6      1          5        1   138
#> [127,]  54   3      0        180        0   128
#> [128,]  56   3      0          8        1   139
#> [129,]  55   1      0          2        0   130
#> [130,]  54   7      1          2        0   129
#> [131,]  52   9      1          3        0   170
#> [132,]  54   2      1          1        1   176
#> [133,]  57   1      0          1        1   100
#> [134,]  56   4      1          0        1   140
#> [135,]  55  11      1          7        0   104
#> [136,]  52  15      1         14        0   130
#> [137,]  56  14      1         11        0   130
#> [138,]  53   3      1          0        1   200
#> [139,]  57  10      0        180        1   170
#> [140,]  58   8      0          8        1   130
#> [141,]  54   5      0        180        1   108
#> [142,]  55   3      1          1        1   156
#> [143,]  57   0      0          0        1   150
#> [144,]  57   4      0        180        1   119
#> [145,]  58   6      1          0        1    90
#> [146,]  53  15      1         10        1   130
#> [147,]  54  17      1          8        1   227
#> [148,]  55  13      0        166        1   140
#> [149,]  56   5      0          5        1   150
#> [150,]  54  23      1          8        0   120
#> [151,]  57   4      1          2        1   185
#> [152,]  53   4      0        147        1   145
#> [153,]  53   7      1          0        1   120
#> [154,]  55   3      1          2        0   140
#> [155,]  55   5      0          5        1   131
#> [156,]  54   7      1          0        1   141
#> [157,]  59  15      1         10        0   140
#> [158,]  58   1      1          1        1   200
#> [159,]  55   5      1          0        0   140
#> [160,]  55   2      0          2        0   106
#> [161,]  57   1      0        180        0   148
#> [162,]  60  11      1          9        0   106
#> [163,]  59   3      0        180        0   120
#> [164,]  58   4      1          0        1   160
#> [165,]  57   2      0          2        1   120
#> [166,]  60   5      1          1        0   138
#> [167,]  57   5      0        180        1   130
#> [168,]  57  10      1          9        0   103
#> [169,]  59   6      1          0        1   140
#> [170,]  59   4      1          0        1   152
#> [171,]  58  26      1          0        1   189
#> [172,]  61   9      0          9        1   160
#> [173,]  58   4      1          3        0   120
#> [174,]  59   2      1          1        0   140
#> [175,]  58  14      1          6        0   190
#> [176,]  61   4      1          3        0   151
#> [177,]  61   9      1          8        0   150
#> [178,]  61   3      1          2        1   102
#> [179,]  58   1      0          1        1   100
#> [180,]  57   2      1          0        1   116
#> [181,]  58  10      0         10        1   150
#> [182,]  57   4      1          3        0   138
#> [183,]  57  11      0        180        1   150
#> [184,]  56  18      1         11        1   165
#> [185,]  59   9      1          0        1    80
#> [186,]  55   4      1          3        1   160
#> [187,]  58  11      0        172        1   135
#> [188,]  60  12      1          0        1   114
#> [189,]  55   9      1          7        1   135
#> [190,]  61   4      1          0        1   115
#> [191,]  61  13      1         12        1   130
#> [192,]  59  11      1          8        1   190
#> [193,]  57   1      0          1        0   126
#> [194,]  57  15      1         13        1   110
#> [195,]  59   5      1          2        0   182
#> [196,]  58   5      1          1        1   135
#> [197,]  59  10      0        180        0   160
#> [198,]  61   8      0         77        0   120
#> [199,]  61  13      0         13        0   210
#> [200,]  62  10      1          0        1   153
#> [201,]  57   3      1          0        0   100
#> [202,]  61  28      1          7        0   133
#> [203,]  57   7      0        169        0   180
#> [204,]  60   7      0          7        0   147
#> [205,]  59  13      1          2        0   198
#> [206,]  57  12      1          9        1   120
#> [207,]  62   4      1          0        0   160
#> [208,]  60  17      1          8        1   140
#> [209,]  58   2      0         30        0   202
#> [210,]  59   1      0        180        0   155
#> [211,]  63   6      0         28        1   120
#> [212,]  61  13      0         13        0   120
#> [213,]  61   5      0          5        1   110
#> [214,]  61   5      0          5        1   160
#> [215,]  58  11      1          9        0   179
#> [216,]  57   2      1          1        0   159
#> [217,]  62   1      1          0        1   172
#> [218,]  63   3      1          1        0   180
#> [219,]  63   1      0        180        1   130
#> [220,]  63   4      1          3        0   222
#> [221,]  62   3      0        180        1   105
#> [222,]  63  15      1         10        1   126
#> [223,]  64   4      0        180        0   130
#> [224,]  60  18      1         13        0   132
#> [225,]  59   8      0        180        1   140
#> [226,]  61   9      1          9        1   150
#> [227,]  58   9      1          9        0   110
#> [228,]  62   7      0          7        0   150
#> [229,]  59   1      0         22        1   162
#> [230,]  59   4      0        180        0   196
#> [231,]  60   7      1          5        1   141
#> [232,]  59   5      1          1        0   148
#> [233,]  65  13      0        180        1   100
#> [234,]  63   1      0          1        0   162
#> [235,]  63   1      0          1        0   130
#> [236,]  62   6      0        180        0   170
#> [237,]  61  15      1         13        0   170
#> [238,]  59   4      0          4        0   149
#> [239,]  60   3      0          3        0   168
#> [240,]  64  10      1          9        0   160
#> [241,]  62   6      0          6        0   120
#> [242,]  63  12      1         10        0   200
#> [243,]  59  10      0        180        1   130
#> [244,]  60   8      0         17        1   130
#> [245,]  64  12      1         11        0   160
#> [246,]  66   1      1          0        1   120
#> [247,]  63  10      1          0        1   148
#> [248,]  63  14      1          9        0   123
#> [249,]  65  36      1         11        0   140
#> [250,]  63   4      1          3        0   162
#> [251,]  66   3      1          1        0   127
#> [252,]  64  32      1          9        1   160
#> [253,]  63  12      1          9        0   114
#> [254,]  63   7      0        180        0   120
#> [255,]  66   5      1          0        1   110
#> [256,]  65   8      1          0        0   168
#> [257,]  65  10      1          8        1   120
#> [258,]  64   0      0          0        1    90
#> [259,]  60   6      0        180        0   130
#> [260,]  64  21      1         10        0   190
#> [261,]  64   9      0        180        0   150
#> [262,]  61   4      0        180        1   113
#> [263,]  65   3      0        180        1   190
#> [264,]  64   7      0        180        1   120
#> [265,]  63  12      0         12        1   150
#> [266,]  62   3      1          1        1   199
#> [267,]  65   6      0          9        0   112
#> [268,]  65   3      1          0        1    80
#> [269,]  63   5      1          4        0   170
#> [270,]  63   2      1          1        0   180
#> [271,]  62  13      1         11        0   180
#> [272,]  67  11      0         11        1   100
#> [273,]  64   2      0          2        0   201
#> [274,]  66  18      1          5        0   142
#> [275,]  62   9      0        180        0   145
#> [276,]  61  14      1          5        0   140
#> [277,]  63   2      1          0        0   140
#> [278,]  64  19      1          8        1   160
#> [279,]  64  13      1         12        1   150
#> [280,]  66   7      1          0        1   115
#> [281,]  66  13      1          0        0   118
#> [282,]  64  14      1         13        1   150
#> [283,]  65   3      0          3        0   105
#> [284,]  64   0      0          0        1   148
#> [285,]  66   3      1          0        1   135
#> [286,]  66   6      1          0        1   140
#> [287,]  68   1      0        180        1   166
#> [288,]  64  10      1          9        1   110
#> [289,]  63   7      1          0        0   162
#> [290,]  67   8      1          1        1   130
#> [291,]  63  10      0         16        1   160
#> [292,]  66  14      0        180        0   130
#> [293,]  63   8      1          1        1   162
#> [294,]  63  10      0         18        1   130
#> [295,]  67  11      0         11        0   150
#> [296,]  68  11      0        180        0   160
#> [297,]  68  14      0         79        0   172
#> [298,]  65  15      1         12        1   150
#> [299,]  66  11      1          0        0   100
#> [300,]  69  12      0         15        1   140
#> [301,]  63   2      0        180        0   150
#> [302,]  69   6      0        180        1   100
#> [303,]  66   9      1          8        0   130
#> [304,]  65   8      1          0        1    90
#> [305,]  66   3      0          3        1   138
#> [306,]  69   1      1          0        0   170
#> [307,]  67   1      0        180        1   160
#> [308,]  68  10      1         10        1   150
#> [309,]  65   1      1          0        0   133
#> [310,]  67   7      1          4        1   130
#> [311,]  63   2      1          0        0    99
#> [312,]  65  10      1          1        1   148
#> [313,]  66  19      1         12        1   150
#> [314,]  67  12      1         12        0   160
#> [315,]  69   6      0         99        1   140
#> [316,]  65   4      1          1        0   130
#> [317,]  64   4      0        179        0   160
#> [318,]  66   4      0        180        1   130
#> [319,]  70  15      1         12        1   132
#> [320,]  64  11      0         11        0   125
#> [321,]  64   4      0        180        1   140
#> [322,]  64   0      1          0        1   118
#> [323,]  67   2      0         18        0   131
#> [324,]  66   7      1          5        1   131
#> [325,]  66   4      0        180        0   177
#> [326,]  68   4      1          0        1   160
#> [327,]  69   4      1          3        1   150
#> [328,]  65  13      1         12        1   130
#> [329,]  64  21      0         21        1   155
#> [330,]  66   6      0        180        0   140
#> [331,]  65   1      0          1        1   120
#> [332,]  65   6      0        101        1   115
#> [333,]  68   7      0        150        0   210
#> [334,]  71  20      1          0        1   160
#> [335,]  66   9      1          3        1   151
#> [336,]  66   1      1          1        1   165
#> [337,]  70  14      0        171        0   166
#> [338,]  66   4      0        180        0   130
#> [339,]  69   0      0          0        1   148
#> [340,]  65   2      0        180        0   130
#> [341,]  68   7      1          0        1   150
#> [342,]  67  14      1         13        0   130
#> [343,]  65  14      1         13        1   150
#> [344,]  66   2      0          2        1   228
#> [345,]  71   6      0         45        1   158
#> [346,]  69   5      0          5        1   142
#> [347,]  71   3      0        103        0   133
#> [348,]  70  22      1         13        0   103
#> [349,]  67   1      0         36        1   104
#> [350,]  68   6      0        180        0   145
#> [351,]  69   8      1          5        1   195
#> [352,]  69   6      1          4        1   174
#> [353,]  72   3      1          0        1   132
#> [354,]  67   3      0        180        0   110
#> [355,]  68  18      0         18        1   100
#> [356,]  69  11      1          0        1   120
#> [357,]  66   2      0        180        0   130
#> [358,]  67   7      1          4        0   122
#> [359,]  69   4      1          3        0   132
#> [360,]  69   8      1          2        0   121
#> [361,]  67  13      1          9        0   130
#> [362,]  70   3      0        123        0   130
#> [363,]  67  22      1          1        1   140
#> [364,]  67  12      1          8        0   120
#> [365,]  69   1      0          1        1   110
#> [366,]  67   1      0          1        1    60
#> [367,]  67   4      0         60        1   136
#> [368,]  69   5      0         76        0   120
#> [369,]  67   8      1          0        1   130
#> [370,]  72  13      1         11        1   195
#> [371,]  68  10      1          8        1   160
#> [372,]  72  30      1          0        1   145
#> [373,]  70   7      0          7        0   102
#> [374,]  68   7      1          2        0   135
#> [375,]  73  20      1          0        1   170
#> [376,]  71   6      0          9        0   120
#> [377,]  70  11      0        180        1   210
#> [378,]  72  19      1          8        0   120
#> [379,]  72  12      1         10        0   170
#> [380,]  67   9      0        180        0   158
#> [381,]  72   2      0          2        1   100
#> [382,]  72   6      1          5        0   115
#> [383,]  69   3      0        180        0   220
#> [384,]  72   5      0         28        0   120
#> [385,]  73   6      0        180        1   117
#> [386,]  69   8      1          1        0   164
#> [387,]  68   7      0        180        1   130
#> [388,]  69   1      1          0        0   155
#> [389,]  73   6      1          0        1   270
#> [390,]  71   2      1          0        1   180
#> [391,]  73   7      0          7        1   140
#> [392,]  68  15      1         13        1   130
#> [393,]  70   3      0          3        1   159
#> [394,]  72   6      0        180        1   130
#> [395,]  73   0      0        180        1   161
#> [396,]  74   8      1          0        1    85
#> [397,]  69   2      1          0        1   110
#> [398,]  74  20      0         20        1   180
#> [399,]  68   9      0        180        1   120
#> [400,]  73   3      1          0        1   136
#> [401,]  70   5      1          0        1   190
#> [402,]  71   8      1          7        0   149
#> [403,]  71   3      1          2        1   190
#> [404,]  73  10      1          8        0   106
#> [405,]  69  12      1          1        1   149
#> [406,]  74   4      0          4        0   120
#> [407,]  73   4      0         58        1   160
#> [408,]  72   5      1          3        1   160
#> [409,]  70   3      0        180        1   154
#> [410,]  73   6      0        180        0   110
#> [411,]  72  15      1          0        1   150
#> [412,]  74   3      0          3        1   150
#> [413,]  71  13      1          8        0   121
#> [414,]  69   2      1          1        1    80
#> [415,]  74   7      1          0        1   117
#> [416,]  72  10      1          8        1   153
#> [417,]  69   7      0        180        1   144
#> [418,]  72  15      1         13        0   156
#> [419,]  70   8      0          8        0   120
#> [420,]  75   2      1          1        0   145
#> [421,]  73  10      1          9        1   146
#> [422,]  72  10      1          9        1   160
#> [423,]  73  10      1         10        1   120
#> [424,]  74  15      1          9        1   179
#> [425,]  75   9      1          7        0   140
#> [426,]  75  13      1          1        1   130
#> [427,]  71  11      1          8        0   110
#> [428,]  71   4      0          4        0   134
#> [429,]  72  15      1         12        1   120
#> [430,]  73  10      1          8        0   120
#> [431,]  70   7      1          4        0   184
#> [432,]  72   7      0         57        1   145
#> [433,]  73  10      0        180        0   162
#> [434,]  72  11      0         11        1   140
#> [435,]  73   5      1          3        1   112
#> [436,]  73  12      1         12        1   140
#> [437,]  75   1      0        180        1   140
#> [438,]  72   4      1          0        1   197
#> [439,]  73   5      0        180        0   126
#> [440,]  74  34      1          8        1   233
#> [441,]  76   3      1          0        1   120
#> [442,]  71  32      1         12        1   107
#> [443,]  72   5      0        180        0   154
#> [444,]  72   3      0        180        0   160
#> [445,]  76   5      0          5        1   130
#> [446,]  77  11      0         11        1   150
#> [447,]  77   4      0          4        0   185
#> [448,]  75   3      1          1        0   180
#> [449,]  71  16      0        180        0   140
#> [450,]  73  10      1         10        0   124
#> [451,]  74   3      0          3        1   128
#> [452,]  76   1      0        180        0   114
#> [453,]  72   4      0         85        1   120
#> [454,]  72   4      1          3        0   160
#> [455,]  76  17      1          0        1   200
#> [456,]  73   4      1          3        1   125
#> [457,]  76  13      1         10        0   110
#> [458,]  75   4      1          0        1   122
#> [459,]  75   7      0          7        0   190
#> [460,]  75   0      0          0        1   130
#> [461,]  75  12      0         12        1   160
#> [462,]  74   8      1          0        1   105
#> [463,]  76  13      1          8        1   148
#> [464,]  75   4      1          2        1   188
#> [465,]  74   6      0        180        0   160
#> [466,]  76   4      0          4        1   155
#> [467,]  75   1      0          1        1   125
#> [468,]  74   2      0        180        0   111
#> [469,]  73   1      0         52        1   105
#> [470,]  72   5      0        180        0   120
#> [471,]  76  44      1         10        0   105
#> [472,]  76   5      1          0        1   167
#> [473,]  74   8      1          8        1   170
#> [474,]  75   9      0        180        1   140
#> [475,]  77  12      1          9        1   100
#> [476,]  77  12      0        180        0   130
#> [477,]  78   5      1          0        1   170
#> [478,]  74   6      0         79        1   140
#> [479,]  79  10      1          8        0   190
#> [480,]  74   2      1          0        1   130
#> [481,]  78   7      0          7        1   133
#> [482,]  74  15      0        180        1   172
#> [483,]  74   7      0          7        0   161
#> [484,]  76  13      1          1        1   170
#> [485,]  79   6      0        180        0   170
#> [486,]  78   0      0        180        1   212
#> [487,]  78  13      1          5        0   130
#> [488,]  75   5      0        119        1   150
#> [489,]  75  12      1          1        1   120
#> [490,]  75  13      1          6        0   150
#> [491,]  76   1      0          1        1    83
#> [492,]  78  12      1          9        0   150
#> [493,]  78   2      1          1        0   130
#> [494,]  75   4      1          0        0   212
#> [495,]  77   2      1          0        1   143
#> [496,]  78  10      0        180        1   130
#> [497,]  76  11      1          0        0   120
#> [498,]  75  11      1          4        0   162
#> [499,]  77  24      0         24        1   160
#> [500,]  79   8      0         32        1   120
#> [501,]  80   6      0          6        1   150
#> [502,]  78   6      1          0        1   240
#> [503,]  78  11      1          1        1   140
#> [504,]  76   4      0          4        1   160
#> [505,]  79   4      0          4        1   125
#> [506,]  76  12      1         10        1   127
#> [507,]  77   6      0          6        1   107
#> [508,]  75   2      1          1        1   204
#> [509,]  78  11      0        180        1   135
#> [510,]  76   1      0          1        1   140
#> [511,]  76   1      0          1        1    90
#> [512,]  78   7      1          0        1   110
#> [513,]  79   3      0          3        0   120
#> [514,]  77   6      0          6        1   144
#> [515,]  79   4      1          0        1   120
#> [516,]  81   1      0        180        0   120
#> [517,]  80  15      1         12        1   150
#> [518,]  77   9      1          4        0   141
#> [519,]  82   5      0          8        1   120
#> [520,]  80  40      1          0        1   138
#> [521,]  78   4      0         59        1   112
#> [522,]  80  17      1         12        0   100
#> [523,]  76   7      0        161        0   151
#> [524,]  80  15      1          0        1    90
#> [525,]  81   4      1          2        1   126
#> [526,]  80   9      0        118        1   186
#> [527,]  80   6      0        173        1   160
#> [528,]  78  32      0        180        1   130
#> [529,]  81   3      0        180        0   184
#> [530,]  78   7      0          7        1   147
#> [531,]  78  15      0         15        0   165
#> [532,]  78  26      1          5        0   194
#> [533,]  76   1      0        166        0   131
#> [534,]  81   4      1          1        1   104
#> [535,]  78  20      1          0        1   109
#> [536,]  80   1      0          1        0   100
#> [537,]  78   3      1          1        1   152
#> [538,]  77  10      1          8        1   130
#> [539,]  82   3      1          1        1   144
#> [540,]  77   5      0         85        0   188
#> [541,]  80   2      1          1        0   168
#> [542,]  79   6      0          6        0   152
#> [543,]  80   6      1          0        1   119
#> [544,]  80   5      0          5        1   130
#> [545,]  82   1      1          0        1    82
#> [546,]  81   1      0        108        0   129
#> [547,]  79   1      0        125        0   193
#> [548,]  82  21      1          2        0   155
#> [549,]  84  22      1         10        0   180
#> [550,]  79   4      0          4        1   121
#> [551,]  80   6      0          6        1   110
#> [552,]  82   5      0        180        0   110
#> [553,]  79   7      1          6        0   130
#> [554,]  81  11      1          8        0   160
#> [555,]  81   5      0        177        0    41
#> [556,]  80  11      1          8        0   170
#> [557,]  78  23      1         10        1   145
#> [558,]  78   9      1          4        1   120
#> [559,]  79   1      0        180        1   170
#> [560,]  81  15      0        180        1   140
#> [561,]  80   7      1          0        1   146
#> [562,]  84   5      1          1        1    85
#> [563,]  81  20      1          9        0   170
#> [564,]  83   8      0          8        0   115
#> [565,]  81  16      0         16        1   110
#> [566,]  80  11      1          8        0   110
#> [567,]  81   8      0        180        0   146
#> [568,]  79   7      0        177        0   197
#> [569,]  85   4      0        180        0    90
#> [570,]  81   2      1          1        0   198
#> [571,]  83   2      0          2        1   155
#> [572,]  82   6      0        128        1   100
#> [573,]  84   4      0        167        0   198
#> [574,]  80   3      1          1        1   230
#> [575,]  82  23      1          0        0   110
#> [576,]  84   4      0          4        1    85
#> [577,]  81   1      0          1        1   150
#> [578,]  84   1      0         38        1   205
#> [579,]  83   3      0        180        0   174
#> [580,]  81   4      0         90        1   138
#> [581,]  79   9      1          8        0   150
#> [582,]  80  13      1          8        1   140
#> [583,]  80   2      1          0        1   130
#> [584,]  80   6      0         71        1   189
#> [585,]  83   1      0          1        1   100
#> [586,]  80  30      1         13        0   220
#> [587,]  83   9      0        180        0   198
#> [588,]  79  14      1          0        0   110
#> [589,]  83   3      0        114        0    98
#> [590,]  85   9      1          6        1   160
#> [591,]  81   4      0          4        0   175
#> [592,]  84  15      1         13        1   110
#> [593,]  81  12      0         12        1   163
#> [594,]  82  16      1          8        0   103
#> [595,]  81   4      0          4        0   160
#> [596,]  81  19      1         14        0   120
#> [597,]  82  15      1          0        0   183
#> [598,]  80   2      0         88        0   135
#> [599,]  86   8      0          8        1   132
#> [600,]  81  16      1          9        0   180
#> [601,]  84   6      0        165        0   145
#> [602,]  86   3      0          3        1   140
#> [603,]  82   9      0        180        1   134
#> [604,]  84   3      0        180        1   120
#> [605,]  81   2      1          0        1   118
#> [606,]  83   9      0        180        1   149
#> [607,]  82   1      0        180        1   193
#> [608,]  87   2      0          5        1   137
#> [609,]  82  14      1         11        1   103
#> [610,]  86   6      1          0        1   140
#> [611,]  84  16      0         70        1   150
#> [612,]  86   2      0        180        1   169
#> [613,]  84   3      0          3        1   121
#> [614,]  83  13      1         12        0   170
#> [615,]  87   2      0        180        0   113
#> [616,]  84   9      0         92        1   110
#> [617,]  82   4      0          4        0   130
#> [618,]  86  13      0        177        0   163
#> [619,]  85   3      0          3        1   113
#> [620,]  86   6      0          6        1   117
#> [621,]  88   4      0          4        0   100
#> [622,]  83  20      1          3        1   150
#> [623,]  85  22      0         22        1   184
#> [624,]  83   9      0         65        1   150
#> [625,]  86   9      1          7        1   142
#> [626,]  88   3      0        115        0   110
#> [627,]  87   8      0          8        1   157
#> [628,]  88   4      0          4        0    86
#> [629,]  89   5      0        119        1   140
#> [630,]  87   6      0        180        1   110
#> [631,]  87   1      0          1        0   170
#> [632,]  84   8      0        180        1   119
#> [633,]  85   8      0          8        1   136
#> [634,]  84   2      0        110        1   174
#> [635,]  87  29      0         29        1    97
#> [636,]  84   0      0        180        1   136
#> [637,]  90  14      0         14        1   100
#> [638,]  86   4      0        180        1   145
#> [639,]  87   2      0        180        0   160
#> [640,]  91  10      0        145        0   135
#> [641,]  88   7      0         24        0   119
#> [642,]  88   8      0         50        1   154
#> [643,]  90  11      1         10        1   186
#> [644,]  87   6      0        126        1   168
#> [645,]  86  10      0        180        1   137
#> [646,]  86   9      1          7        0   130
#> [647,]  90   4      1          0        0   121
#> [648,]  87  43      0        178        1   130
#> [649,]  87   5      0         36        1   150
#> [650,]  88   3      1          2        0   159
#> [651,]  92   1      0          1        1   167
#> [652,]  88   5      0        158        0   100
#> [653,]  89  12      1          0        1   130
#> [654,]  89   2      0        168        0   118
#> [655,]  89  52      0         52        1   130
#> [656,]  90  19      1         11        1   129
#> [657,]  94   6      0         50        0    78
#> [658,]  90   1      0          1        1   118
#> [659,]  91   2      0          2        1   116
#> [660,]  94   8      0          8        1   142
#> [661,]  92   4      0         76        1   149
#> [662,]  90  16      0         16        1   106
#> [663,]  96   3      0         12        1    97
#> [664,]  94   3      0         26        1   144
#> [665,]  91  12      0         53        1   212
#> [666,]  91   7      0          7        0   135
#> [667,]  92   5      0         69        0   139
#> [668,]  93   4      0        180        1   135
#> [669,]  96   3      1          0        1   104
#> [670,]  96  15      1          0        1   140
#> 
#> $y
#>   [1] 180.0+   2.0+   5.0+ 180.0+ 180.0+ 180.0+ 115.0  180.0+ 180.0+  12.0 
#>  [11] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 180.0+   5.0+   2.0+
#>  [21] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 180.0+ 155.0+
#>  [31] 180.0+ 180.0+ 180.0+   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 150.0 
#>  [41] 180.0+ 180.0+ 180.0+ 180.0+   6.0+ 180.0+ 180.0+ 180.0+ 180.0+  73.0 
#>  [51] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [61] 180.0+ 180.0+ 180.0+ 180.0+  10.0+ 180.0+ 180.0+ 180.0+ 180.0+   7.0 
#>  [71] 180.0+   1.0  179.0+ 180.0+ 179.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [81] 180.0+ 180.0+   7.0   88.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0  180.0+
#>  [91] 180.0+ 180.0+ 180.0+ 180.0+ 166.0+  99.0   16.0+ 180.0+ 152.0+   7.0+
#> [101]   6.0+ 180.0+  13.0+ 171.0+ 180.0+ 180.0+ 174.0+  28.0    6.0+   1.0 
#> [111] 180.0+   9.0+ 180.0+ 180.0+ 180.0+ 175.0+ 180.0+ 180.0+   7.0+ 180.0+
#> [121] 180.0+ 180.0+ 180.0+  16.0  180.0+ 180.0+ 180.0+   8.0    2.0  180.0+
#> [131] 180.0+ 180.0+   1.0  165.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+   8.0+
#> [141] 180.0+ 180.0+   0.5  180.0+ 180.0+ 180.0+ 171.0+ 166.0+   5.0+ 180.0+
#> [151]   4.0+ 147.0+ 180.0+ 180.0+   5.0+ 180.0+ 180.0+   1.0  180.0+   2.0+
#> [161] 180.0+ 180.0+ 180.0+ 180.0+   2.0  180.0+ 180.0+ 180.0+  64.0  180.0+
#> [171] 180.0+   9.0+ 180.0+ 180.0+ 171.0+ 180.0+ 180.0+   3.0    1.0  180.0+
#> [181]  10.0+ 180.0+ 180.0+ 180.0+   9.0+ 180.0+ 172.0+ 172.0+  24.0  180.0+
#> [191] 180.0+ 180.0+   1.0+  15.0  180.0+ 180.0+ 180.0+  77.0   13.0+ 180.0+
#> [201] 180.0+  94.0  169.0    7.0+ 180.0+ 180.0+ 180.0+ 180.0+  30.0  180.0+
#> [211]  28.0   13.0+   5.0    5.0+ 180.0+ 180.0+   1.0  180.0+ 180.0+ 180.0+
#> [221] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   9.0    7.0+  22.0  180.0+
#> [231]  84.0  180.0+ 180.0+   1.0    1.0  180.0+ 180.0+   4.0+   3.0+ 167.0 
#> [241]   6.0+ 180.0+ 180.0+  17.0   12.0  180.0+ 180.0+  14.0+  36.0  180.0+
#> [251]   3.0+ 180.0+  12.0  180.0+ 180.0+ 180.0+ 180.0+   0.5  180.0+ 180.0+
#> [261] 180.0+ 180.0+ 180.0+ 180.0+  12.0  180.0+   9.0    3.0  180.0+ 180.0+
#> [271] 180.0+  11.0+   2.0+  18.0+ 180.0+ 180.0+   2.0+ 103.0   13.0  179.0+
#> [281] 166.0+  14.0+   3.0    0.5+   3.0+ 180.0+ 180.0+ 180.0+   7.0+   8.0 
#> [291]  16.0  180.0+ 180.0+  18.0   11.0+ 180.0+  79.0   15.0+ 180.0+  15.0 
#> [301] 180.0+ 180.0+ 180.0+   8.0+   3.0  175.0  180.0+  10.0  180.0+ 180.0+
#> [311] 180.0+ 180.0+  19.0+  12.0   99.0  180.0+ 179.0+ 180.0+ 180.0+  11.0+
#> [321] 180.0+   0.5   18.0    7.0+ 180.0+ 180.0+ 152.0+ 180.0+  21.0+ 180.0+
#> [331]   1.0  101.0  150.0  180.0+ 180.0+   1.0  171.0  180.0+   0.5  180.0+
#> [341] 180.0+ 180.0+  14.0+   2.0   45.0    5.0+ 103.0  180.0+  36.0  180.0+
#> [351] 180.0+  97.0  180.0+ 180.0+  18.0  180.0+ 180.0+   7.0  180.0+   8.0+
#> [361]  13.0+ 123.0   51.0  180.0+   1.0    1.0   60.0   76.0  180.0+ 132.0 
#> [371]  10.0+ 162.0    7.0+   7.0+ 124.0    9.0  180.0+ 180.0+  12.0  180.0+
#> [381]   2.0  180.0+ 180.0+  28.0  180.0+ 180.0+ 180.0+ 180.0+   6.0  180.0+
#> [391]   7.0+  15.0    3.0+ 180.0+ 180.0+ 180.0+   2.0   20.0  180.0+ 180.0+
#> [401] 180.0+   8.0    3.0   87.0   12.0    4.0+  58.0  180.0+ 180.0+ 180.0+
#> [411] 180.0+   3.0  175.0    2.0  180.0+  10.0+ 180.0+ 180.0+   8.0+ 180.0+
#> [421] 180.0+ 159.0   15.0  180.0+ 180.0+  13.0  180.0+   4.0+ 180.0+  10.0 
#> [431] 104.0+  57.0  180.0+  11.0    5.0   12.0  180.0+ 180.0+ 180.0+  34.0 
#> [441] 180.0+ 177.0+ 180.0+ 180.0+   5.0   11.0+   4.0+ 180.0+ 180.0+  10.0 
#> [451]   3.0  180.0+  85.0  180.0+  17.0+ 180.0+ 174.0+   4.0    7.0    0.5 
#> [461]  12.0  180.0+ 180.0+  46.0  180.0+   4.0    1.0  180.0+  52.0  180.0+
#> [471] 180.0+ 180.0+   8.0  180.0+ 180.0+ 180.0+ 180.0+  79.0  180.0+ 176.0+
#> [481]   7.0  180.0+   7.0  180.0+ 180.0+ 180.0+ 172.0  119.0   12.0  180.0+
#> [491]   1.0  180.0+ 180.0+   4.0+   2.0  180.0+  11.0  152.0+  24.0   32.0 
#> [501]   6.0  180.0+ 180.0+   4.0    4.0  180.0+   6.0    2.0+ 180.0+   1.0 
#> [511]   1.0   43.0    3.0    6.0  138.0  180.0+ 180.0+  71.0    8.0   40.0 
#> [521]  59.0   17.0  161.0  180.0+  93.0  118.0  173.0  180.0+ 180.0+   7.0+
#> [531]  15.0+ 171.0+ 166.0+  71.0   20.0+   1.0    3.0+  10.0  180.0+  85.0 
#> [541]  10.0    6.0+   6.0    5.0    1.0  108.0  125.0  180.0+ 180.0+   4.0 
#> [551]   6.0  180.0+ 180.0+ 180.0+ 177.0+ 169.0   70.0  180.0+ 180.0+ 180.0+
#> [561]   7.0+ 180.0+  20.0    8.0+  16.0  180.0+ 180.0+ 177.0+ 180.0+ 180.0+
#> [571]   2.0  128.0  167.0    3.0+  62.0    4.0    1.0   38.0  180.0+  90.0 
#> [581] 180.0+ 180.0+ 180.0+  71.0    1.0   30.0  180.0+ 180.0+ 114.0  180.0+
#> [591]   4.0+ 180.0+  12.0   16.0+   4.0+ 180.0+  83.0   88.0    8.0  180.0+
#> [601] 165.0    3.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0  174.0    6.0 
#> [611]  70.0  180.0+   3.0   13.0  180.0+  92.0    4.0  177.0    3.0+   6.0+
#> [621]   4.0+  20.0   22.0   65.0   11.0  115.0    8.0+   4.0  119.0  180.0+
#> [631]   1.0+ 180.0+   8.0  110.0   29.0  180.0+  14.0  180.0+ 180.0+ 145.0 
#> [641]  24.0   50.0   11.0  126.0  180.0+ 180.0+   4.0  178.0+  36.0   75.0 
#> [651]   1.0  158.0  180.0+ 168.0   52.0  180.0+  50.0    1.0+   2.0    8.0+
#> [661]  76.0   16.0   12.0   26.0   53.0    7.0+  69.0  180.0+   3.0   15.0+
#> 
#> $weights
#> NULL
#> 
#> $offset
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Warning: Multiple lambdas have been fit. Lambda will be set to 0.01 (see parameter 's').

# Score the predictions
predictions$score()
#> surv.cindex 
#>    0.853112