GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.coxnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.coxnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Caution: This learner is different to learners calling glmnet::cv.glmnet() in that it does not use the internal optimization of parameter lambda. Instead, lambda needs to be tuned by the user (e.g., via mlr3tuning). When lambda is tuned, the glmnet will be trained for each tuning iteration. While fitting the whole path of lambdas would be more efficient, as is done by default in glmnet::glmnet(), tuning/selecting the parameter at prediction time (using parameter s) is currently not supported in mlr3 (at least not in efficient manner). Tuning the s parameter is, therefore, currently discouraged.

When the data are i.i.d. and efficiency is key, we recommend using the respective auto-tuning counterpart in mlr_learners_surv.cv_glmnet(). However, in some situations this is not applicable, usually when data are imbalanced or not i.i.d. (longitudinal, time-series) and tuning requires custom resampling strategies (blocked design, stratification).

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.coxnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exact	logical	FALSE	TRUE, FALSE	-

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	TRUE	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	0.01		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.logistic	character	Newton	Newton, modified.Newton	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvGlmnet

Methods

Public methods

LearnerSurvGlmnet$new()

LearnerSurvGlmnet$selected_features()

LearnerSurvGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.glmnet")
print(learner)
#> 
#> ── <LearnerSurvGlmnet> (surv.glmnet): Regularized Generalized Linear Model ─────
#> • Model: -
#> • Parameters: use_pred_offset=TRUE
#> • Packages: mlr3, mlr3proba, mlr3extralearners, and glmnet
#> • Predict Types: [crank], distr, and lp
#> • Feature Types: logical, integer, and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: offset, selected_features, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#>    Df  %Dev   Lambda
#> 1   0  0.00 0.183300
#> 2   2  0.63 0.167000
#> 3   2  1.62 0.152200
#> 4   2  2.47 0.138700
#> 5   2  3.20 0.126400
#> 6   3  5.59 0.115100
#> 7   3  7.88 0.104900
#> 8   3  9.61 0.095590
#> 9   3 10.96 0.087100
#> 10  3 12.04 0.079360
#> 11  3 12.92 0.072310
#> 12  3 13.65 0.065890
#> 13  4 14.26 0.060030
#> 14  4 14.78 0.054700
#> 15  5 15.23 0.049840
#> 16  6 15.64 0.045410
#> 17  6 15.99 0.041380
#> 18  6 16.28 0.037700
#> 19  6 16.53 0.034350
#> 20  6 16.74 0.031300
#> 21  6 16.91 0.028520
#> 22  6 17.06 0.025990
#> 23  6 17.19 0.023680
#> 24  6 17.29 0.021570
#> 25  6 17.38 0.019660
#> 26  6 17.45 0.017910
#> 27  6 17.52 0.016320
#> 28  6 17.57 0.014870
#> 29  6 17.61 0.013550
#> 30  6 17.65 0.012350
#> 31  6 17.68 0.011250
#> 32  6 17.71 0.010250
#> 33  6 17.73 0.009339
#> 34  6 17.75 0.008510
#> 35  6 17.76 0.007754
#> 36  6 17.77 0.007065
#> 37  6 17.78 0.006437
#> 38  6 17.79 0.005865
#> 39  6 17.80 0.005344
#> 40  6 17.81 0.004870
#> 41  6 17.81 0.004437
#> 42  6 17.82 0.004043
#> 43  6 17.82 0.003684
#> 44  6 17.82 0.003356
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  32   5      1          0        1   121
#>   [2,]  33   2      0          2        0   150
#>   [3,]  35   5      1          2        0   172
#>   [4,]  35  10      1          9        0   106
#>   [5,]  34   5      0          5        0   120
#>   [6,]  35   2      1          1        1   112
#>   [7,]  37   9      0        180        1   151
#>   [8,]  38  13      1          0        1   161
#>   [9,]  36   1      0        180        1   155
#>  [10,]  38  12      1          8        1   120
#>  [11,]  36   5      1          0        1   115
#>  [12,]  33   6      1          1        1   115
#>  [13,]  38  16      1         10        0   160
#>  [14,]  38  12      1         11        1    92
#>  [15,]  40  12      1          9        0   153
#>  [16,]  37   1      1          0        1   146
#>  [17,]  40   2      1          1        1   148
#>  [18,]  38   5      1          3        0   125
#>  [19,]  40   6      0        180        1   138
#>  [20,]  40  11      1         10        1   120
#>  [21,]  42   2      0        180        0   100
#>  [22,]  43   3      1          0        1   100
#>  [23,]  41   2      1          1        0   166
#>  [24,]  40   1      1          0        1   145
#>  [25,]  43   4      1          0        1   130
#>  [26,]  42  15      1         13        1   125
#>  [27,]  40   3      1          1        0   170
#>  [28,]  43   2      1          1        1   116
#>  [29,]  44   5      1          1        0   170
#>  [30,]  45   3      0        180        1   154
#>  [31,]  41  13      1          1        0   140
#>  [32,]  45   9      1          7        0   110
#>  [33,]  45   6      0        180        1   170
#>  [34,]  41   5      1          4        1   141
#>  [35,]  44   2      1          1        1   150
#>  [36,]  43   2      0        180        1   140
#>  [37,]  45   2      0        180        1   140
#>  [38,]  46  15      0        180        0   120
#>  [39,]  46   2      1          1        0   126
#>  [40,]  47   4      1          3        0   118
#>  [41,]  48  15      0        180        1   160
#>  [42,]  45   3      0        150        0   130
#>  [43,]  46   7      1          2        0   166
#>  [44,]  43  29      0        180        1   180
#>  [45,]  43  10      0        180        0   185
#>  [46,]  47   4      1          3        1   160
#>  [47,]  45   8      1          0        1   117
#>  [48,]  49   5      0         73        1   136
#>  [49,]  46   6      1          0        1   100
#>  [50,]  44   4      1          0        1   114
#>  [51,]  47   2      0        180        0   108
#>  [52,]  45   5      0        180        1   190
#>  [53,]  46   5      1          3        0   130
#>  [54,]  46   4      0        180        1   121
#>  [55,]  46  15      0        180        1   120
#>  [56,]  45   9      1          0        1   145
#>  [57,]  48   3      0        180        0   154
#>  [58,]  47   5      1          3        1   130
#>  [59,]  47   9      1          6        0   170
#>  [60,]  46   3      1          0        1   119
#>  [61,]  49   4      0        180        0   117
#>  [62,]  47  10      0         10        1   140
#>  [63,]  50   1      1          0        1   129
#>  [64,]  48   2      1          0        0   184
#>  [65,]  47   7      0        180        0   145
#>  [66,]  49   7      1          7        1   110
#>  [67,]  46   3      1          1        1   140
#>  [68,]  46   9      1          9        1   122
#>  [69,]  50   7      0        180        1   110
#>  [70,]  51   1      0          1        1   145
#>  [71,]  49  15      1         11        1   160
#>  [72,]  47   2      0        180        0   150
#>  [73,]  49  23      0        179        1   112
#>  [74,]  46   6      1          0        1   156
#>  [75,]  52   2      0        180        1   170
#>  [76,]  50   7      1          0        1    92
#>  [77,]  50   4      0          4        1   100
#>  [78,]  50   1      1          0        0   150
#>  [79,]  49   7      1          4        1    90
#>  [80,]  52   2      0        180        0   155
#>  [81,]  46   3      0        180        1   120
#>  [82,]  46   1      1          1        0   145
#>  [83,]  50   4      1          1        0   150
#>  [84,]  53   8      0         36        1   160
#>  [85,]  47   2      1          1        0   110
#>  [86,]  53   5      0        180        1   140
#>  [87,]  54  17      1         12        1   102
#>  [88,]  53   5      0         77        0   159
#>  [89,]  54   6      1          3        0   129
#>  [90,]  51   3      1          1        0   140
#>  [91,]  50   2      0          5        1   106
#>  [92,]  50  14      1         13        0   170
#>  [93,]  48   3      1          2        0   150
#>  [94,]  49   5      1          2        1   150
#>  [95,]  53   4      0          4        0   140
#>  [96,]  52  14      1          7        1   200
#>  [97,]  53   4      1          0        1   156
#>  [98,]  51  13      0         99        1   160
#>  [99,]  54   9      1          0        1   138
#> [100,]  49  16      0         16        0   125
#> [101,]  55   3      1          1        0   150
#> [102,]  54  23      1         10        0   131
#> [103,]  55   6      1          2        1   114
#> [104,]  54   9      1          1        0   130
#> [105,]  55   4      1          2        0   150
#> [106,]  52   4      0        180        1   180
#> [107,]  52   4      0        180        0   183
#> [108,]  55  28      1         13        1   160
#> [109,]  49   6      1          0        1   130
#> [110,]  49   1      0          1        1   110
#> [111,]  53   9      0          9        1    95
#> [112,]  53   8      1          0        1   130
#> [113,]  56   4      1          1        1   130
#> [114,]  54   7      1          0        1   100
#> [115,]  56   2      0        180        0   132
#> [116,]  52   8      0        180        0   119
#> [117,]  54   3      0        180        1   180
#> [118,]  55   6      0        180        0   170
#> [119,]  52  16      0         16        0   152
#> [120,]  53   4      0        180        1   150
#> [121,]  55   6      0        180        1   100
#> [122,]  54  12      1          0        1   190
#> [123,]  56   3      0          8        1   139
#> [124,]  57   3      0          3        0   120
#> [125,]  54   7      1          2        0   129
#> [126,]  54   2      1          1        0   135
#> [127,]  57   1      0        180        1   156
#> [128,]  57   1      0          1        1   100
#> [129,]  52   2      0        180        0   140
#> [130,]  55  11      1          7        0   104
#> [131,]  52  15      1         14        0   130
#> [132,]  56  14      1         11        0   130
#> [133,]  53   3      1          0        1   200
#> [134,]  57  10      0        180        1   170
#> [135,]  54   5      0        180        1   108
#> [136,]  55   3      1          1        1   156
#> [137,]  57   0      0          0        1   150
#> [138,]  53  21      1         13        1   130
#> [139,]  57   4      0        180        1   119
#> [140,]  58   6      1          0        1    90
#> [141,]  53  15      1         10        1   130
#> [142,]  55  13      0        166        1   140
#> [143,]  56   5      0          5        1   150
#> [144,]  53   4      0        147        1   145
#> [145,]  53   7      1          0        1   120
#> [146,]  54   7      1          0        1   141
#> [147,]  56   4      0          4        0   164
#> [148,]  58   9      1          0        1   180
#> [149,]  58   1      1          1        1   200
#> [150,]  56   7      1          5        1   120
#> [151,]  55   2      0          2        0   106
#> [152,]  59   9      1          1        1   125
#> [153,]  57   1      0        180        0   148
#> [154,]  60  11      1          9        0   106
#> [155,]  59   3      0        180        0   120
#> [156,]  60   5      1          1        0   138
#> [157,]  57   5      0        180        1   130
#> [158,]  55   5      1          0        1   160
#> [159,]  57  10      1          9        0   103
#> [160,]  59   4      1          0        1   152
#> [161,]  61   9      0          9        1   160
#> [162,]  60   0      1          0        1    80
#> [163,]  59   2      1          1        0   140
#> [164,]  58   8      0        161        1   140
#> [165,]  61   4      1          3        0   151
#> [166,]  61   3      1          2        1   102
#> [167,]  58   1      0          1        1   100
#> [168,]  57  13      1         10        0   110
#> [169,]  57   4      1          3        0   138
#> [170,]  61   3      0         17        0   143
#> [171,]  56  14      0         45        0   130
#> [172,]  57   3      1          2        0   120
#> [173,]  58  19      1         13        1   140
#> [174,]  56  13      1          6        1   158
#> [175,]  56  18      1         11        1   165
#> [176,]  55   4      1          3        1   160
#> [177,]  58  11      0        172        1   135
#> [178,]  61   4      1          0        1   115
#> [179,]  56   8      1          8        0   120
#> [180,]  59  11      1          8        1   190
#> [181,]  57   1      0          1        0   126
#> [182,]  59   5      1          2        0   182
#> [183,]  58   5      1          1        1   135
#> [184,]  59  10      0        180        0   160
#> [185,]  61   8      0         77        0   120
#> [186,]  61  13      0         13        0   210
#> [187,]  62  10      1          0        1   153
#> [188,]  62   7      1          2        1   180
#> [189,]  57   3      1          0        0   100
#> [190,]  61  18      0        170        0   140
#> [191,]  61  28      1          7        0   133
#> [192,]  58   8      1          3        1   150
#> [193,]  60   7      0          7        0   147
#> [194,]  59  13      1          2        0   198
#> [195,]  57  12      1          9        1   120
#> [196,]  60  17      1          8        1   140
#> [197,]  58   2      0         30        0   202
#> [198,]  59   1      0        180        0   155
#> [199,]  59  16      1          9        1   133
#> [200,]  63   6      0         28        1   120
#> [201,]  57  18      1          9        1    93
#> [202,]  58  11      1          9        0   179
#> [203,]  57   2      1          1        0   159
#> [204,]  62   1      1          0        1   172
#> [205,]  58   7      0        180        1   150
#> [206,]  63   1      0        180        1   130
#> [207,]  61   7      0        180        0   135
#> [208,]  63   4      0        180        1   190
#> [209,]  64   4      0        180        0   130
#> [210,]  63   4      1          1        0   155
#> [211,]  60  18      1         13        0   132
#> [212,]  59   8      0        180        1   140
#> [213,]  61   9      1          9        1   150
#> [214,]  58   9      1          9        0   110
#> [215,]  62   7      0          7        0   150
#> [216,]  59   1      0         22        1   162
#> [217,]  58   2      0        180        0   127
#> [218,]  59   4      0        180        0   196
#> [219,]  60   7      1          5        1   141
#> [220,]  60   7      0          7        0   140
#> [221,]  60   7      1          1        1    90
#> [222,]  63   1      0          1        0   130
#> [223,]  62   6      0        180        0   170
#> [224,]  61  15      1         13        0   170
#> [225,]  60   3      0          3        0   168
#> [226,]  64  10      1          9        0   160
#> [227,]  62   6      0          6        0   120
#> [228,]  63  12      1         10        0   200
#> [229,]  59  10      0        180        1   130
#> [230,]  60   8      0         17        1   130
#> [231,]  61   6      1          1        1   117
#> [232,]  64  12      1         11        0   160
#> [233,]  64   6      1          0        1   140
#> [234,]  63  10      1          0        1   148
#> [235,]  63   4      1          3        0   162
#> [236,]  66   3      1          1        0   127
#> [237,]  61  10      1          2        1   194
#> [238,]  63  12      1          9        0   114
#> [239,]  66   5      1          0        1   110
#> [240,]  65  10      1          8        1   120
#> [241,]  64   0      0          0        1    90
#> [242,]  64  21      1         10        0   190
#> [243,]  61  12      1         11        0   154
#> [244,]  64   9      0        180        0   150
#> [245,]  61   4      0        180        1   113
#> [246,]  65   3      0        180        1   190
#> [247,]  63  16      1          7        1   110
#> [248,]  63  12      0         12        1   150
#> [249,]  65   6      0          9        0   112
#> [250,]  63   2      1          1        0   180
#> [251,]  62  13      1         11        0   180
#> [252,]  67  11      0         11        1   100
#> [253,]  64   2      0          2        0   201
#> [254,]  66  18      1          5        0   142
#> [255,]  66  16      1         11        1   169
#> [256,]  62   9      0        180        0   145
#> [257,]  61  15      1         10        0   130
#> [258,]  63   9      1          8        1   160
#> [259,]  63   3      1          2        0   120
#> [260,]  63   2      1          0        0   140
#> [261,]  64  19      1          8        1   160
#> [262,]  67   6      0        180        1   170
#> [263,]  68   5      1          4        1   150
#> [264,]  64  13      1         12        1   150
#> [265,]  66   7      1          0        1   115
#> [266,]  64  14      1         13        1   150
#> [267,]  64   0      0          0        1   148
#> [268,]  66   3      1          0        1   135
#> [269,]  66   6      1          0        1   140
#> [270,]  65   2      1          1        1   170
#> [271,]  68   1      0        180        1   166
#> [272,]  64  10      1          9        1   110
#> [273,]  63   7      1          0        0   162
#> [274,]  67   8      1          1        1   130
#> [275,]  68   5      0          5        1    90
#> [276,]  64   1      0          1        1   120
#> [277,]  63   8      1          1        1   162
#> [278,]  65  18      1          3        0   120
#> [279,]  63  10      0         18        1   130
#> [280,]  68  11      0        180        0   160
#> [281,]  68  14      0         79        0   172
#> [282,]  66  11      1          0        0   100
#> [283,]  63   2      0        180        0   150
#> [284,]  65  11      1          6        0   130
#> [285,]  69   6      0        180        1   100
#> [286,]  66   9      1          8        0   130
#> [287,]  63   8      0        180        1   120
#> [288,]  68  14      1         13        1   140
#> [289,]  68  10      1         10        1   150
#> [290,]  65   1      1          0        0   133
#> [291,]  63   2      1          0        0    99
#> [292,]  65   6      0          6        0    80
#> [293,]  66  19      1         12        1   150
#> [294,]  67  12      1         12        0   160
#> [295,]  65   4      1          1        0   130
#> [296,]  64   4      0        179        0   160
#> [297,]  70  15      1         12        1   132
#> [298,]  64  11      0         11        0   125
#> [299,]  64   4      0        180        1   140
#> [300,]  64   0      1          0        1   118
#> [301,]  67   2      0         18        0   131
#> [302,]  66   7      1          5        1   131
#> [303,]  68   4      1          0        1   160
#> [304,]  65  13      1         12        1   130
#> [305,]  69   8      0         93        0   140
#> [306,]  65   1      0          1        1   120
#> [307,]  68  18      1          0        1   160
#> [308,]  68   4      0          4        1   190
#> [309,]  71   3      0          5        0   112
#> [310,]  70   7      1          0        1   190
#> [311,]  68   7      0        150        0   210
#> [312,]  71  20      1          0        1   160
#> [313,]  67   2      0        180        0   128
#> [314,]  66   9      1          3        1   151
#> [315,]  66   1      1          1        1   165
#> [316,]  70   4      1          0        1   180
#> [317,]  66   4      0        180        0   130
#> [318,]  67  10      1          9        0   200
#> [319,]  67   6      1          4        0   130
#> [320,]  65   2      0        180        0   130
#> [321,]  68   7      1          0        1   150
#> [322,]  69   3      1          2        0   151
#> [323,]  67  14      1         13        0   130
#> [324,]  71   7      0          7        0   230
#> [325,]  66   2      0          2        1   228
#> [326,]  71   6      0         45        1   158
#> [327,]  69   3      0          3        1   130
#> [328,]  70  22      1         13        0   103
#> [329,]  67   1      0         36        1   104
#> [330,]  67   5      0          5        0   130
#> [331,]  68   6      0        180        0   145
#> [332,]  69   6      1          4        1   174
#> [333,]  72   3      1          0        1   132
#> [334,]  72   7      0          7        1   110
#> [335,]  67   3      0        180        0   110
#> [336,]  66   2      1          1        0   123
#> [337,]  69  19      0        180        0   130
#> [338,]  67  14      0        172        1   140
#> [339,]  66   2      0        180        0   130
#> [340,]  67   7      1          4        0   122
#> [341,]  69   4      1          3        0   132
#> [342,]  68   2      0          7        1   130
#> [343,]  69   8      1          2        0   121
#> [344,]  72   5      1          4        0   170
#> [345,]  68   3      0         19        0   135
#> [346,]  67   4      0         60        1   136
#> [347,]  69   5      0         76        0   120
#> [348,]  67   8      1          0        1   130
#> [349,]  72  13      1         11        1   195
#> [350,]  68  10      1          8        1   160
#> [351,]  66  24      1         13        0   130
#> [352,]  70  35      1          0        1   105
#> [353,]  68   7      1          2        0   135
#> [354,]  73  20      1          0        1   170
#> [355,]  70  11      0        180        1   210
#> [356,]  72  19      1          8        0   120
#> [357,]  72  12      1         10        0   170
#> [358,]  67   5      1          0        1   147
#> [359,]  73  13      0        152        1   130
#> [360,]  70   5      0        180        0   150
#> [361,]  72   2      0          2        1   100
#> [362,]  67   4      1          1        0   134
#> [363,]  72   6      1          5        0   115
#> [364,]  71   1      0        173        1   188
#> [365,]  69   3      0        180        0   220
#> [366,]  68   4      1          3        0   210
#> [367,]  72   5      0         28        0   120
#> [368,]  71   5      0        180        0   191
#> [369,]  73   6      0        180        1   117
#> [370,]  69  16      1         10        1   140
#> [371,]  69   8      1          1        0   164
#> [372,]  72  16      1          1        1   130
#> [373,]  70   4      0        180        0   180
#> [374,]  69   1      1          0        0   155
#> [375,]  73   6      1          0        1   270
#> [376,]  72   8      1          1        1   150
#> [377,]  71   2      1          0        1   180
#> [378,]  70   3      0          3        1   159
#> [379,]  70  13      1          9        0   100
#> [380,]  74   8      1          0        1    85
#> [381,]  73   4      0        180        1   154
#> [382,]  69   2      1          0        1   110
#> [383,]  71  15      1         11        0   165
#> [384,]  74  20      0         20        1   180
#> [385,]  68   9      0        180        1   120
#> [386,]  74   0      1          0        1    90
#> [387,]  73   3      1          0        1   136
#> [388,]  71  17      1         11        0   160
#> [389,]  71   8      1          7        0   149
#> [390,]  71   3      1          2        1   190
#> [391,]  73  10      1          8        0   106
#> [392,]  69  12      1          1        1   149
#> [393,]  74   4      0          4        0   120
#> [394,]  73   4      0         58        1   160
#> [395,]  70   3      0        180        1   154
#> [396,]  73   6      0        180        0   110
#> [397,]  71   7      1          2        0   143
#> [398,]  72   8      1          0        1   140
#> [399,]  74   3      0          3        1   150
#> [400,]  73  17      1         11        0   140
#> [401,]  71  13      1          8        0   121
#> [402,]  70   4      1          0        1   140
#> [403,]  72  10      1          8        1   153
#> [404,]  69   7      0        180        1   144
#> [405,]  71  10      1          9        1   120
#> [406,]  75   1      0          1        0   133
#> [407,]  75   2      1          1        0   145
#> [408,]  73  10      1          9        1   146
#> [409,]  74  15      1          9        1   179
#> [410,]  73   1      0          1        1    80
#> [411,]  75   9      1          7        0   140
#> [412,]  75  13      1          1        1   130
#> [413,]  73  10      1          8        0   120
#> [414,]  70   7      1          4        0   184
#> [415,]  72   1      1          1        0   168
#> [416,]  72   7      0         57        1   145
#> [417,]  73  10      0        180        0   162
#> [418,]  72  11      0         11        1   140
#> [419,]  70   3      0          3        0   150
#> [420,]  73   5      1          3        1   112
#> [421,]  76  25      1         12        1   170
#> [422,]  73  12      1         12        1   140
#> [423,]  72   2      0        180        0   120
#> [424,]  75   1      0        180        1   140
#> [425,]  71   3      1          0        0   144
#> [426,]  73   5      0        180        0   126
#> [427,]  73   4      0        180        0   124
#> [428,]  74  34      1          8        1   233
#> [429,]  76   3      1          0        1   120
#> [430,]  71  32      1         12        1   107
#> [431,]  72   3      0        180        0   160
#> [432,]  77   4      0          4        0   185
#> [433,]  75   3      1          1        0   180
#> [434,]  72   7      1          2        0   142
#> [435,]  73  15      0         15        1   160
#> [436,]  71  16      0        180        0   140
#> [437,]  73  10      1         10        0   124
#> [438,]  74   7      0        180        1   150
#> [439,]  76   1      0        180        0   114
#> [440,]  74   2      1          1        0   140
#> [441,]  76   8      1          0        1   141
#> [442,]  74  19      1          4        1   200
#> [443,]  73   6      0          6        1   114
#> [444,]  74   2      0        180        0   190
#> [445,]  72   4      0         85        1   120
#> [446,]  72   4      1          3        0   160
#> [447,]  73   4      1          3        1   125
#> [448,]  75   0      0          0        1   130
#> [449,]  73  13      1         11        0   195
#> [450,]  75   4      1          2        1   188
#> [451,]  74   6      0        180        0   160
#> [452,]  74   2      0        180        0   111
#> [453,]  73   1      0         52        1   105
#> [454,]  73   0      0        180        0   156
#> [455,]  72   5      0        180        0   120
#> [456,]  78  12      1         11        1   160
#> [457,]  76  44      1         10        0   105
#> [458,]  74   8      1          8        1   170
#> [459,]  75   9      0        180        1   140
#> [460,]  77   5      1          0        0   123
#> [461,]  77  12      1          9        1   100
#> [462,]  77   1      1          0        1    90
#> [463,]  78   5      1          0        1   170
#> [464,]  73   7      1          0        0   174
#> [465,]  74   6      0         79        1   140
#> [466,]  75   3      1          1        1   171
#> [467,]  74   9      1          8        0   200
#> [468,]  74   2      1          0        1   130
#> [469,]  77   3      0        180        0   110
#> [470,]  73   8      1          1        1   162
#> [471,]  78   7      0          7        1   133
#> [472,]  74  15      0        180        1   172
#> [473,]  78   8      1          6        1   110
#> [474,]  76  13      1          1        1   170
#> [475,]  78  32      1          9        1   198
#> [476,]  79   6      0        180        0   170
#> [477,]  78  13      1          5        0   130
#> [478,]  75   5      0        119        1   150
#> [479,]  80   8      0          8        1   120
#> [480,]  74  10      1          8        0   135
#> [481,]  76   1      0          1        1    83
#> [482,]  79   4      0         80        0   145
#> [483,]  78  12      1          9        0   150
#> [484,]  78   2      1          1        0   130
#> [485,]  78  10      0        180        1   130
#> [486,]  76  11      1          0        0   120
#> [487,]  76   7      0         29        1   150
#> [488,]  79   8      0         32        1   120
#> [489,]  80   9      0         23        1   128
#> [490,]  80   6      0          6        1   150
#> [491,]  78   6      1          0        1   240
#> [492,]  79  11      0        180        0   160
#> [493,]  79   2      1          0        1   121
#> [494,]  78  14      1          0        1   140
#> [495,]  78  11      1          8        1   118
#> [496,]  76   4      0          4        1   160
#> [497,]  76  10      1          8        0   180
#> [498,]  76  12      1         10        1   127
#> [499,]  77   6      0          6        1   107
#> [500,]  80   3      1          0        1   120
#> [501,]  75   2      1          1        1   204
#> [502,]  78  11      0        180        1   135
#> [503,]  76   1      0          1        1   140
#> [504,]  76   1      0          1        1    90
#> [505,]  78   7      1          0        1   110
#> [506,]  79   3      0          3        0   120
#> [507,]  77   7      0        180        1   170
#> [508,]  77   6      0          6        1   144
#> [509,]  79   4      1          0        1   120
#> [510,]  80  15      1         12        1   150
#> [511,]  77   9      1          4        0   141
#> [512,]  82   5      0          8        1   120
#> [513,]  78   4      0         59        1   112
#> [514,]  80  17      1         12        0   100
#> [515,]  79  10      0         10        1   120
#> [516,]  80  15      1          0        1    90
#> [517,]  81   4      1          2        1   126
#> [518,]  80   9      0        118        1   186
#> [519,]  79   1      0         37        1   140
#> [520,]  81   3      0        180        0   184
#> [521,]  81   2      0        175        0   172
#> [522,]  78   7      0          7        1   147
#> [523,]  77  13      1          0        1   190
#> [524,]  78  15      0         15        0   165
#> [525,]  80   5      1          1        1   108
#> [526,]  78   4      0        180        0   175
#> [527,]  79   3      0          3        1   101
#> [528,]  78  26      1          5        0   194
#> [529,]  76   1      0        166        0   131
#> [530,]  81   4      1          1        1   104
#> [531,]  78  20      1          0        1   109
#> [532,]  80   1      0          1        0   100
#> [533,]  78   3      1          1        1   152
#> [534,]  77  10      1          8        1   130
#> [535,]  82   3      1          1        1   144
#> [536,]  80   2      1          1        0   168
#> [537,]  79   6      0          6        0   152
#> [538,]  78   2      0        180        0   148
#> [539,]  82   1      1          0        1    82
#> [540,]  79  10      0        180        1   150
#> [541,]  77   4      0        180        1    98
#> [542,]  81   1      0        108        0   129
#> [543,]  79   1      0        125        0   193
#> [544,]  82  21      1          2        0   155
#> [545,]  84  22      1         10        0   180
#> [546,]  83   9      1          5        1   170
#> [547,]  82   5      0        180        0   110
#> [548,]  83   5      0        180        0   148
#> [549,]  79   7      1          6        0   130
#> [550,]  83   4      0        103        0    97
#> [551,]  81  11      1          8        0   160
#> [552,]  80  11      1          8        0   170
#> [553,]  78  23      1         10        1   145
#> [554,]  79   1      0        180        1   170
#> [555,]  81  15      0        180        1   140
#> [556,]  80   7      1          0        1   146
#> [557,]  81  20      1          9        0   170
#> [558,]  83   8      0          8        0   115
#> [559,]  81  16      0         16        1   110
#> [560,]  80   6      1          0        1   150
#> [561,]  80  11      1          8        0   110
#> [562,]  81   8      0        180        0   146
#> [563,]  80   8      1          7        0   160
#> [564,]  79   7      0        177        0   197
#> [565,]  79   0      1          0        1    96
#> [566,]  85   4      0        180        0    90
#> [567,]  81   2      1          1        0   198
#> [568,]  83   2      0          2        1   155
#> [569,]  82   6      0        128        1   100
#> [570,]  80   3      1          1        1   230
#> [571,]  82  23      1          0        0   110
#> [572,]  84   5      0        180        1   203
#> [573,]  81   1      0          1        1   150
#> [574,]  84   1      0         38        1   205
#> [575,]  83   3      0        180        0   174
#> [576,]  85   3      1          2        1   160
#> [577,]  80  13      1          8        1   140
#> [578,]  79   4      0          4        1    60
#> [579,]  80   6      0         71        1   189
#> [580,]  83   1      0          1        1   100
#> [581,]  79  14      1          0        0   110
#> [582,]  83   3      0        114        0    98
#> [583,]  81  14      1         12        1   128
#> [584,]  83   2      0        154        0   130
#> [585,]  81   4      0          4        0   175
#> [586,]  84  15      1         13        1   110
#> [587,]  81   1      0          1        1   145
#> [588,]  81   4      0          4        0   160
#> [589,]  86  12      0        180        1   120
#> [590,]  83  12      1          2        1   170
#> [591,]  82  15      1          0        0   183
#> [592,]  83   7      0        126        0   135
#> [593,]  86   8      0          8        1   132
#> [594,]  84   6      0        165        0   145
#> [595,]  86   3      0          3        1   140
#> [596,]  82   9      0        180        1   134
#> [597,]  84   3      0        180        1   120
#> [598,]  81   2      1          0        1   118
#> [599,]  81   4      0        180        0   160
#> [600,]  82   1      0        180        1   193
#> [601,]  87   2      0          5        1   137
#> [602,]  86   6      1          0        1   140
#> [603,]  84   3      1          2        0   125
#> [604,]  83  10      1          0        1   190
#> [605,]  86   2      0        180        1   169
#> [606,]  84   3      0          3        1   121
#> [607,]  83  13      1         12        0   170
#> [608,]  84   7      1          2        0   148
#> [609,]  87   2      0        180        0   113
#> [610,]  84   9      0         92        1   110
#> [611,]  84   3      0        180        1   170
#> [612,]  86   4      0         38        1   122
#> [613,]  82   4      0          4        0   130
#> [614,]  86  13      0        177        0   163
#> [615,]  86   6      0          6        1   117
#> [616,]  84  13      0         62        1   100
#> [617,]  88   4      0          4        1   115
#> [618,]  83   9      0         65        1   150
#> [619,]  86   9      1          7        1   142
#> [620,]  87   2      0        180        1   130
#> [621,]  88   3      0        115        0   110
#> [622,]  88   2      0        180        1    68
#> [623,]  83   3      0          3        1   130
#> [624,]  86  15      1          8        1   109
#> [625,]  89   5      0        119        1   140
#> [626,]  84   2      0        110        1   174
#> [627,]  87  29      0         29        1    97
#> [628,]  87  15      1          9        1   138
#> [629,]  84   0      0        180        1   136
#> [630,]  88   1      0          1        0   135
#> [631,]  86   4      0        180        1   145
#> [632,]  87   2      0        180        0   160
#> [633,]  87   6      1          0        0   125
#> [634,]  91  10      0        145        0   135
#> [635,]  86   3      1          0        1    80
#> [636,]  88   7      0         24        0   119
#> [637,]  90  11      1         10        1   186
#> [638,]  87   6      0        126        1   168
#> [639,]  86  10      0        180        1   137
#> [640,]  91   1      0          1        1    74
#> [641,]  90   5      1          0        1   125
#> [642,]  89   3      1          1        1   160
#> [643,]  92   1      0          1        1   167
#> [644,]  91   3      0         33        1   137
#> [645,]  88   5      0        158        0   100
#> [646,]  89  12      1          0        1   130
#> [647,]  89   2      0        168        0   118
#> [648,]  91   5      0        169        1   176
#> [649,]  89  52      0         52        1   130
#> [650,]  92   7      0          7        1   110
#> [651,]  89  14      0        180        1    84
#> [652,]  90  18      0        180        0   188
#> [653,]  91   4      1          0        1   120
#> [654,]  94   6      0         50        0    78
#> [655,]  90   1      0          1        1   118
#> [656,]  91   2      0          2        1   116
#> [657,]  93   8      0        179        1   110
#> [658,]  92   4      0         76        1   149
#> [659,]  91   1      0        180        0   158
#> [660,]  90  16      0         16        1   106
#> [661,]  96   3      0         12        1    97
#> [662,]  95   8      1          5        1   150
#> [663,]  94   3      0         26        1   144
#> [664,]  91  12      0         53        1   212
#> [665,]  93   0      1          0        1   122
#> [666,]  92   5      0         69        0   139
#> [667,]  92   2      0          2        0   112
#> [668,]  93   4      0        180        1   135
#> [669,]  96   3      1          0        1   104
#> [670,]  96  15      1          0        1   140
#> 
#> $y
#>   [1]   5.0+   2.0+   5.0+ 180.0+   5.0+   2.0+ 180.0+ 180.0+ 180.0+  12.0 
#>  [11]   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+   5.0+ 180.0+ 180.0+
#>  [21] 180.0+   3.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 155.0+ 180.0+
#>  [31] 180.0+ 180.0+ 180.0+   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [41] 180.0+ 150.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+  73.0  180.0+ 180.0+
#>  [51] 180.0+ 180.0+   5.0+ 180.0+ 180.0+ 177.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [61] 180.0+  10.0+ 172.0+ 180.0+ 180.0+   7.0  180.0+ 180.0+ 180.0+   1.0 
#>  [71] 179.0+ 180.0+ 179.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+ 180.0+ 180.0+
#>  [81] 180.0+ 180.0+ 180.0+  36.0  180.0+ 180.0+ 180.0+  77.0  180.0+ 180.0+
#>  [91]   5.0  180.0+ 180.0+ 180.0+   4.0+  85.0  166.0+  99.0  180.0+  16.0+
#> [101] 180.0+ 152.0+   6.0+ 180.0+ 180.0+ 180.0+ 180.0+  28.0    6.0+   1.0 
#> [111]   9.0+ 180.0+ 180.0+   7.0+ 180.0+ 180.0+ 180.0+ 180.0+  16.0+ 180.0+
#> [121] 180.0+  12.0+   8.0    3.0+ 180.0+ 180.0+ 180.0+   1.0  180.0+ 180.0+
#> [131] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   0.5  180.0+ 180.0+ 180.0+
#> [141] 180.0+ 166.0+   5.0+ 147.0+ 180.0+ 180.0+   4.0+   9.0+   1.0  180.0+
#> [151]   2.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [161]   9.0+   0.5  180.0+ 161.0+ 180.0+   3.0    1.0  180.0+ 180.0+  17.0 
#> [171]  45.0    3.0+  19.0  180.0+ 180.0+ 180.0+ 172.0+ 180.0+   8.0  180.0+
#> [181]   1.0+ 180.0+ 180.0+ 180.0+  77.0   13.0+ 180.0+ 180.0+ 180.0+ 170.0 
#> [191]  94.0  180.0+   7.0+ 180.0+ 180.0+ 180.0+  30.0  180.0+ 180.0+  28.0 
#> [201]  18.0  180.0+ 180.0+   1.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0+
#> [211] 180.0+ 180.0+ 180.0+   9.0    7.0+  22.0  180.0+ 180.0+  84.0    7.0+
#> [221] 180.0+   1.0  180.0+ 180.0+   3.0+ 167.0    6.0+ 180.0+ 180.0+  17.0 
#> [231] 180.0+  12.0  180.0+ 180.0+ 180.0+   3.0+  88.0   12.0  180.0+ 180.0+
#> [241]   0.5  180.0+  12.0+ 180.0+ 180.0+ 180.0+ 180.0+  12.0    9.0  180.0+
#> [251] 180.0+  11.0+   2.0+  18.0+ 180.0+ 180.0+ 180.0+ 180.0+   3.0+   2.0+
#> [261] 103.0  180.0+   5.0+  13.0  179.0+  14.0+   0.5+   3.0+ 180.0+ 175.0+
#> [271] 180.0+ 180.0+   7.0+   8.0    5.0    1.0  180.0+ 123.0+  18.0  180.0+
#> [281]  79.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  10.0  180.0+
#> [291] 180.0+   6.0   19.0+  12.0  180.0+ 179.0+ 180.0+  11.0+ 180.0+   0.5 
#> [301]  18.0    7.0+ 180.0+ 180.0+  93.0    1.0   18.0+   4.0    5.0    7.0+
#> [311] 150.0  180.0+ 180.0+ 180.0+   1.0  180.0+ 180.0+ 174.0+   6.0  180.0+
#> [321] 180.0+ 180.0+ 180.0+   7.0+   2.0   45.0    3.0+ 180.0+  36.0    5.0+
#> [331] 180.0+  97.0  180.0+   7.0  180.0+   2.0+ 180.0+ 172.0+ 180.0+   7.0 
#> [341] 180.0+   7.0    8.0+ 180.0+  19.0   60.0   76.0  180.0+ 132.0   10.0+
#> [351] 180.0+ 180.0+   7.0+ 124.0  180.0+ 180.0+  12.0  180.0+ 152.0  180.0+
#> [361]   2.0   76.0  180.0+ 173.0+ 180.0+ 180.0+  28.0  180.0+ 180.0+  16.0+
#> [371] 180.0+  16.0+ 180.0+ 180.0+   6.0  180.0+ 180.0+   3.0+  13.0+ 180.0+
#> [381] 180.0+   2.0  180.0+  20.0  180.0+   0.5  180.0+ 180.0+   8.0    3.0 
#> [391]  87.0   12.0    4.0+  58.0  180.0+ 180.0+ 180.0+ 180.0+   3.0  180.0+
#> [401] 175.0  180.0+  10.0+ 180.0+ 179.0+   1.0  180.0+ 180.0+ 180.0+   1.0 
#> [411] 180.0+  13.0   10.0  104.0+   1.0   57.0  180.0+  11.0    3.0+   5.0 
#> [421] 180.0+  12.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+  34.0  180.0+ 177.0+
#> [431] 180.0+   4.0+ 180.0+   7.0   15.0+ 180.0+  10.0  180.0+ 180.0+ 180.0+
#> [441] 180.0+ 180.0+   6.0  180.0+  85.0  180.0+ 180.0+   0.5  180.0+  46.0 
#> [451] 180.0+ 180.0+  52.0  180.0+ 180.0+  12.0  180.0+   8.0  180.0+   5.0 
#> [461] 180.0+   1.0  180.0+   7.0+  79.0    3.0  168.0+ 176.0+ 180.0+ 180.0+
#> [471]   7.0  180.0+   8.0+ 180.0+  32.0  180.0+ 172.0  119.0    8.0  180.0+
#> [481]   1.0   80.0  180.0+ 180.0+ 180.0+  11.0   29.0   32.0   23.0    6.0 
#> [491] 180.0+ 180.0+ 180.0+ 180.0+  11.0    4.0   10.0+ 180.0+   6.0    3.0+
#> [501]   2.0+ 180.0+   1.0    1.0   43.0    3.0  180.0+   6.0  138.0  180.0+
#> [511]  71.0    8.0   59.0   17.0   10.0+ 180.0+  93.0  118.0   37.0  180.0+
#> [521] 175.0+   7.0+  22.0   15.0+   5.0+ 180.0+   3.0  171.0+ 166.0+  71.0 
#> [531]  20.0+   1.0    3.0+  10.0  180.0+  10.0    6.0+ 180.0+   1.0  180.0+
#> [541] 180.0+ 108.0  125.0  180.0+ 180.0+   9.0+ 180.0+ 180.0+ 180.0+ 103.0 
#> [551] 180.0+ 169.0   70.0  180.0+ 180.0+   7.0+  20.0    8.0+  16.0  180.0+
#> [561] 180.0+ 180.0+ 180.0+ 177.0+   0.5  180.0+ 180.0+   2.0  128.0    3.0+
#> [571]  62.0  180.0+   1.0   38.0  180.0+ 180.0+ 180.0+   4.0   71.0    1.0 
#> [581] 180.0+ 114.0  180.0+ 154.0    4.0+ 180.0+   1.0    4.0+ 180.0+  77.0 
#> [591]  83.0  126.0    8.0  165.0    3.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [601]   5.0    6.0  180.0+ 180.0+ 180.0+   3.0   13.0  180.0+ 180.0+  92.0 
#> [611] 180.0+  38.0    4.0  177.0    6.0+  62.0    4.0   65.0   11.0  180.0+
#> [621] 115.0  180.0+   3.0+ 180.0+ 119.0  110.0   29.0  180.0+ 180.0+   1.0+
#> [631] 180.0+ 180.0+  25.0  145.0    3.0   24.0   11.0  126.0  180.0+   1.0 
#> [641]  89.0    3.0+   1.0   33.0  158.0  180.0+ 168.0  169.0   52.0    7.0 
#> [651] 180.0+ 180.0+   4.0   50.0    1.0+   2.0  179.0+  76.0  180.0+  16.0 
#> [661]  12.0    8.0   26.0   53.0    0.5   69.0    2.0  180.0+   3.0   15.0+
#> 
#> $weights
#> NULL
#> 
#> $offset
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Warning: Multiple lambdas have been fit. Lambda will be set to 0.01 (see parameter 's').

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8433608