GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.coxnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.coxnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Caution: This learner is different to learners calling glmnet::cv.glmnet() in that it does not use the internal optimization of parameter lambda. Instead, lambda needs to be tuned by the user (e.g., via mlr3tuning). When lambda is tuned, the glmnet will be trained for each tuning iteration. While fitting the whole path of lambdas would be more efficient, as is done by default in glmnet::glmnet(), tuning/selecting the parameter at prediction time (using parameter s) is currently not supported in mlr3 (at least not in efficient manner). Tuning the s parameter is, therefore, currently discouraged.

When the data are i.i.d. and efficiency is key, we recommend using the respective auto-tuning counterpart in mlr_learners_surv.cv_glmnet(). However, in some situations this is not applicable, usually when data are imbalanced or not i.i.d. (longitudinal, time-series) and tuning requires custom resampling strategies (blocked design, stratification).

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.coxnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exact	logical	FALSE	TRUE, FALSE	-

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	TRUE	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	0.01		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.logistic	character	Newton	Newton, modified.Newton	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvGlmnet

Methods

Public methods

LearnerSurvGlmnet$new()

LearnerSurvGlmnet$selected_features()

LearnerSurvGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.glmnet")
print(learner)
#> 
#> ── <LearnerSurvGlmnet> (surv.glmnet): Regularized Generalized Linear Model ─────
#> • Model: -
#> • Parameters: use_pred_offset=TRUE
#> • Packages: mlr3, mlr3proba, mlr3extralearners, and glmnet
#> • Predict Types: [crank], distr, and lp
#> • Feature Types: logical, integer, and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: offset, selected_features, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#>    Df  %Dev   Lambda
#> 1   0  0.00 0.198400
#> 2   1  0.63 0.180700
#> 3   2  1.27 0.164700
#> 4   2  2.12 0.150100
#> 5   2  2.85 0.136700
#> 6   3  4.25 0.124600
#> 7   3  6.74 0.113500
#> 8   3  8.61 0.103400
#> 9   4 10.07 0.094240
#> 10  4 11.30 0.085870
#> 11  4 12.30 0.078240
#> 12  4 13.12 0.071290
#> 13  4 13.81 0.064960
#> 14  4 14.38 0.059190
#> 15  4 14.86 0.053930
#> 16  5 15.27 0.049140
#> 17  5 15.63 0.044770
#> 18  5 15.94 0.040800
#> 19  6 16.22 0.037170
#> 20  6 16.46 0.033870
#> 21  6 16.66 0.030860
#> 22  6 16.83 0.028120
#> 23  6 16.98 0.025620
#> 24  6 17.10 0.023340
#> 25  6 17.20 0.021270
#> 26  6 17.29 0.019380
#> 27  6 17.36 0.017660
#> 28  6 17.42 0.016090
#> 29  6 17.47 0.014660
#> 30  6 17.51 0.013360
#> 31  6 17.55 0.012170
#> 32  6 17.58 0.011090
#> 33  6 17.61 0.010110
#> 34  6 17.63 0.009208
#> 35  6 17.65 0.008390
#> 36  6 17.66 0.007644
#> 37  6 17.67 0.006965
#> 38  6 17.69 0.006346
#> 39  6 17.69 0.005783
#> 40  6 17.70 0.005269
#> 41  6 17.71 0.004801
#> 42  6 17.71 0.004374
#> 43  6 17.72 0.003986
#> 44  6 17.72 0.003632
#> 45  6 17.72 0.003309
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  28   9      0        180        1   107
#>   [2,]  32   5      1          0        1   121
#>   [3,]  35   5      1          2        0   172
#>   [4,]  35  10      1          9        0   106
#>   [5,]  35   2      0        180        0   121
#>   [6,]  38   2      0        115        0   150
#>   [7,]  36   1      0        180        1   155
#>   [8,]  35   0      0        180        1   119
#>   [9,]  38  12      1          8        1   120
#>  [10,]  36   5      1          0        1   115
#>  [11,]  33   6      1          1        1   115
#>  [12,]  38  12      1         11        1    92
#>  [13,]  37   1      1          0        1   146
#>  [14,]  40   2      1          1        1   148
#>  [15,]  42   2      0        180        1   100
#>  [16,]  38   5      1          3        0   125
#>  [17,]  42   2      0          2        0   140
#>  [18,]  40   6      0        180        1   138
#>  [19,]  40  11      1         10        1   120
#>  [20,]  43   3      1          0        1   100
#>  [21,]  41   2      1          1        0   166
#>  [22,]  40   1      1          0        1   145
#>  [23,]  43   4      1          0        1   130
#>  [24,]  40   3      1          1        0   170
#>  [25,]  43   2      1          1        1   116
#>  [26,]  42   2      0        180        1   124
#>  [27,]  44   5      1          1        0   170
#>  [28,]  45   3      0        180        1   154
#>  [29,]  41  10      1          8        0   150
#>  [30,]  41  13      1          1        0   140
#>  [31,]  45   9      1          7        0   110
#>  [32,]  45   6      0        180        1   170
#>  [33,]  41   5      1          4        1   141
#>  [34,]  44   2      1          1        1   150
#>  [35,]  43   2      0        180        1   140
#>  [36,]  45   2      0        180        1   140
#>  [37,]  46  15      0        180        0   120
#>  [38,]  47   4      1          3        0   118
#>  [39,]  48  15      0        180        1   160
#>  [40,]  44   3      1          0        1   180
#>  [41,]  46   7      1          2        0   166
#>  [42,]  43  29      0        180        1   180
#>  [43,]  45   4      1          0        0   124
#>  [44,]  47   6      1          0        1   116
#>  [45,]  46  13      1         10        0   100
#>  [46,]  43   3      1          0        1   124
#>  [47,]  49   5      0         73        1   136
#>  [48,]  45   5      0          5        0   141
#>  [49,]  44   9      1          8        1   135
#>  [50,]  45   5      0        180        1   190
#>  [51,]  46   5      1          3        0   130
#>  [52,]  46   4      0        180        1   121
#>  [53,]  45   9      1          0        1   145
#>  [54,]  48   3      0        180        0   154
#>  [55,]  48  12      1         11        0   200
#>  [56,]  47   5      1          3        1   130
#>  [57,]  47   9      1          6        0   170
#>  [58,]  46   3      1          0        1   119
#>  [59,]  49   4      0        180        0   117
#>  [60,]  50   1      1          0        1   129
#>  [61,]  47   7      0        180        0   145
#>  [62,]  50   4      1          1        0   125
#>  [63,]  50   6      1          2        1   140
#>  [64,]  49   7      1          7        1   110
#>  [65,]  46   3      1          1        1   140
#>  [66,]  46   9      1          9        1   122
#>  [67,]  49   2      0          2        0   105
#>  [68,]  51   1      0          1        1   145
#>  [69,]  49  15      1         11        1   160
#>  [70,]  47   2      0        180        0   150
#>  [71,]  49  23      0        179        1   112
#>  [72,]  46   6      1          0        1   156
#>  [73,]  52   2      0        180        1   170
#>  [74,]  51   3      1          2        0   113
#>  [75,]  50   1      1          0        0   150
#>  [76,]  49   7      1          4        1    90
#>  [77,]  47   8      0        180        0   160
#>  [78,]  47   6      0        180        1   162
#>  [79,]  51   8      0        180        1   140
#>  [80,]  46   3      0        180        1   120
#>  [81,]  46   1      1          1        0   145
#>  [82,]  50   4      1          1        0   150
#>  [83,]  48   7      1          0        1   110
#>  [84,]  53   8      0         36        1   160
#>  [85,]  48  17      1         10        0   111
#>  [86,]  47   2      1          1        0   110
#>  [87,]  52   4      1          4        0   152
#>  [88,]  49  15      0        180        1   160
#>  [89,]  53   5      0        180        1   140
#>  [90,]  54  17      1         12        1   102
#>  [91,]  53   7      1          0        0   199
#>  [92,]  54   6      1          3        0   129
#>  [93,]  50  10      1          6        0   122
#>  [94,]  51  25      1          1        0   202
#>  [95,]  49   5      1          2        1   150
#>  [96,]  52  14      1          7        1   200
#>  [97,]  48   6      0        180        0   160
#>  [98,]  48  11      1         10        0   120
#>  [99,]  53   4      1          0        1   156
#> [100,]  55   3      1          1        0   150
#> [101,]  54  23      1         10        0   131
#> [102,]  52   7      1          2        0   154
#> [103,]  55   6      1          2        1   114
#> [104,]  54   9      1          1        0   130
#> [105,]  52   4      0        180        1   180
#> [106,]  50   5      1          4        1   150
#> [107,]  50   3      0        174        1   153
#> [108,]  55  28      1         13        1   160
#> [109,]  49   6      1          0        1   130
#> [110,]  50   7      1          1        0   156
#> [111,]  53   9      0          9        1    95
#> [112,]  50   7      1          0        1   127
#> [113,]  56   4      1          1        1   130
#> [114,]  54   7      1          0        1   100
#> [115,]  56   2      0        180        0   132
#> [116,]  55   5      1          4        1   120
#> [117,]  52   8      0        180        0   119
#> [118,]  54   3      0        180        1   180
#> [119,]  55   6      0        180        0   170
#> [120,]  52  16      0         16        0   152
#> [121,]  53  10      1          9        0   172
#> [122,]  52  16      1         14        0   170
#> [123,]  53  15      0         15        1    90
#> [124,]  53   4      0        180        1   150
#> [125,]  55   6      0        180        1   100
#> [126,]  55   6      1          5        1   138
#> [127,]  54  12      1          0        1   190
#> [128,]  55   2      0        134        1   140
#> [129,]  54   3      0        180        0   128
#> [130,]  56   3      0          8        1   139
#> [131,]  55   1      0          2        0   130
#> [132,]  57   3      0          3        0   120
#> [133,]  54   7      1          2        0   129
#> [134,]  54   2      1          1        0   135
#> [135,]  54   2      1          1        1   176
#> [136,]  57   1      0        180        1   156
#> [137,]  56   4      1          0        1   140
#> [138,]  55  11      1          7        0   104
#> [139,]  52  15      1         14        0   130
#> [140,]  56  14      1         11        0   130
#> [141,]  57  10      0        180        1   170
#> [142,]  58   8      0          8        1   130
#> [143,]  54   5      0        180        1   108
#> [144,]  55   3      1          1        1   156
#> [145,]  57   0      0          0        1   150
#> [146,]  53  21      1         13        1   130
#> [147,]  57   4      0        180        1   119
#> [148,]  58   6      1          0        1    90
#> [149,]  53  15      1         10        1   130
#> [150,]  54  17      1          8        1   227
#> [151,]  55   9      1          2        1   147
#> [152,]  55  13      0        166        1   140
#> [153,]  56   5      0          5        1   150
#> [154,]  54  23      1          8        0   120
#> [155,]  57   4      1          2        1   185
#> [156,]  53   4      0        147        1   145
#> [157,]  57  11      1         10        1   129
#> [158,]  55   3      1          2        0   140
#> [159,]  55   5      0          5        1   131
#> [160,]  56   4      0          4        0   164
#> [161,]  59  15      1         10        0   140
#> [162,]  58   9      1          0        1   180
#> [163,]  59   3      0        180        0   120
#> [164,]  58   4      1          0        1   160
#> [165,]  57   2      0          2        1   120
#> [166,]  57   5      0        180        1   130
#> [167,]  58  11      1          9        1   124
#> [168,]  55   5      1          0        1   160
#> [169,]  57  10      1          9        0   103
#> [170,]  59   5      0        180        1   155
#> [171,]  58  26      1          0        1   189
#> [172,]  61   9      0          9        1   160
#> [173,]  60   0      1          0        1    80
#> [174,]  59   2      1          1        0   140
#> [175,]  61   4      1          3        0   151
#> [176,]  61   3      1          2        1   102
#> [177,]  58   1      0          1        1   100
#> [178,]  57  13      1         10        0   110
#> [179,]  57   2      1          0        1   116
#> [180,]  58  10      0         10        1   150
#> [181,]  57   4      1          3        0   138
#> [182,]  57  11      0        180        1   150
#> [183,]  61   3      0         17        0   143
#> [184,]  56  14      0         45        0   130
#> [185,]  58  19      1         13        1   140
#> [186,]  56  13      1          6        1   158
#> [187,]  56  18      1         11        1   165
#> [188,]  59   9      1          0        1    80
#> [189,]  55   4      1          3        1   160
#> [190,]  60  12      1          0        1   114
#> [191,]  55   9      1          7        1   135
#> [192,]  61   4      1          0        1   115
#> [193,]  56   8      1          8        0   120
#> [194,]  61  13      1         12        1   130
#> [195,]  59  11      1          8        1   190
#> [196,]  57   1      0          1        0   126
#> [197,]  57  15      1         13        1   110
#> [198,]  59   5      1          2        0   182
#> [199,]  59  10      0        180        0   160
#> [200,]  61   8      0         77        0   120
#> [201,]  58   8      1          5        0   152
#> [202,]  62  10      1          0        1   153
#> [203,]  62   7      1          2        1   180
#> [204,]  57   3      1          0        0   100
#> [205,]  61  18      0        170        0   140
#> [206,]  58   8      1          3        1   150
#> [207,]  61   7      0          7        1   150
#> [208,]  57  12      1          9        1   120
#> [209,]  62   4      1          0        0   160
#> [210,]  60  17      1          8        1   140
#> [211,]  58   3      1          0        1   146
#> [212,]  58   2      0         30        0   202
#> [213,]  59   1      0        180        0   155
#> [214,]  63   6      0         28        1   120
#> [215,]  61  13      0         13        0   120
#> [216,]  61   5      0          5        1   110
#> [217,]  57  18      1          9        1    93
#> [218,]  61   5      0          5        1   160
#> [219,]  57   2      1          1        0   159
#> [220,]  62  17      1         10        1   180
#> [221,]  62   1      1          0        1   172
#> [222,]  58   7      0        180        1   150
#> [223,]  63   1      0        180        1   130
#> [224,]  62   3      0        180        1   105
#> [225,]  63  15      1         10        1   126
#> [226,]  64   4      0        180        0   130
#> [227,]  60  18      1         13        0   132
#> [228,]  59   8      0        180        1   140
#> [229,]  58   9      1          9        0   110
#> [230,]  59   1      0         22        1   162
#> [231,]  58   2      0        180        0   127
#> [232,]  59   4      0        180        0   196
#> [233,]  60   7      1          5        1   141
#> [234,]  60   7      0          7        0   140
#> [235,]  59   5      1          1        0   148
#> [236,]  60   7      1          1        1    90
#> [237,]  65  13      0        180        1   100
#> [238,]  62   6      0        180        0   170
#> [239,]  59   4      0          4        0   149
#> [240,]  60   3      0          3        0   168
#> [241,]  64  10      1          9        0   160
#> [242,]  63  12      1         10        0   200
#> [243,]  60   8      0         17        1   130
#> [244,]  66   1      1          0        1   120
#> [245,]  65  36      1         11        0   140
#> [246,]  63   4      1          3        0   162
#> [247,]  66   3      1          1        0   127
#> [248,]  64  32      1          9        1   160
#> [249,]  63  12      1          9        0   114
#> [250,]  63   7      0        180        0   120
#> [251,]  64  21      1         10        0   190
#> [252,]  61  12      1         11        0   154
#> [253,]  64   9      0        180        0   150
#> [254,]  61   4      0        180        1   113
#> [255,]  65   3      0        180        1   190
#> [256,]  64   7      0        180        1   120
#> [257,]  66   6      1          1        1   130
#> [258,]  65   6      0          9        0   112
#> [259,]  65   3      1          0        1    80
#> [260,]  63   2      1          1        0   180
#> [261,]  67  11      0         11        1   100
#> [262,]  64   2      0          2        0   201
#> [263,]  66  18      1          5        0   142
#> [264,]  62   9      0        180        0   145
#> [265,]  61  14      1          5        0   140
#> [266,]  63   9      1          8        1   160
#> [267,]  63   3      1          2        0   120
#> [268,]  63   2      1          0        0   140
#> [269,]  67   6      0        180        1   170
#> [270,]  65  15      1         11        1   160
#> [271,]  64  13      1         12        1   150
#> [272,]  66   7      1          0        1   115
#> [273,]  66  13      1          0        0   118
#> [274,]  64  14      1         13        1   150
#> [275,]  65   3      0          3        0   105
#> [276,]  64   0      0          0        1   148
#> [277,]  66   3      1          0        1   135
#> [278,]  65   2      1          1        1   170
#> [279,]  64  10      1          9        1   110
#> [280,]  63   7      1          0        0   162
#> [281,]  63  10      0         16        1   160
#> [282,]  66  14      0        180        0   130
#> [283,]  64   1      0          1        1   120
#> [284,]  68  18      0        180        1   260
#> [285,]  65  17      1         14        1   100
#> [286,]  63   8      1          1        1   162
#> [287,]  65  18      1          3        0   120
#> [288,]  63  10      0         18        1   130
#> [289,]  67  11      0         11        0   150
#> [290,]  68  11      0        180        0   160
#> [291,]  68  14      0         79        0   172
#> [292,]  66  12      1         10        1   150
#> [293,]  65  15      1         12        1   150
#> [294,]  66  11      1          0        0   100
#> [295,]  65   4      1          2        1   145
#> [296,]  69  12      0         15        1   140
#> [297,]  69  21      1         10        0   180
#> [298,]  66   9      1          8        0   130
#> [299,]  63   8      0        180        1   120
#> [300,]  68  14      1         13        1   140
#> [301,]  65   8      1          0        1    90
#> [302,]  66   3      0          3        1   138
#> [303,]  67   1      0        180        1   160
#> [304,]  68  10      1         10        1   150
#> [305,]  65   1      1          0        0   133
#> [306,]  67   7      1          4        1   130
#> [307,]  67   2      0        180        0   184
#> [308,]  65   6      0          6        0    80
#> [309,]  65  10      1          1        1   148
#> [310,]  66  19      1         12        1   150
#> [311,]  67  12      1         12        0   160
#> [312,]  64   4      0        179        0   160
#> [313,]  64  11      0         11        0   125
#> [314,]  67   2      0         18        0   131
#> [315,]  66   7      1          5        1   131
#> [316,]  66   4      0        180        0   177
#> [317,]  68   4      1          0        1   160
#> [318,]  69   4      1          3        1   150
#> [319,]  69  17      1         10        0   140
#> [320,]  69   8      0         93        0   140
#> [321,]  64  21      0         21        1   155
#> [322,]  66   6      0        180        0   140
#> [323,]  65   1      0          1        1   120
#> [324,]  65   6      0        101        1   115
#> [325,]  68   4      0          4        1   190
#> [326,]  71  20      1          0        1   160
#> [327,]  67   2      0        180        0   128
#> [328,]  66   1      1          1        1   165
#> [329,]  70   4      1          0        1   180
#> [330,]  69   8      0        180        1   153
#> [331,]  70  14      0        171        0   166
#> [332,]  66   4      0        180        0   130
#> [333,]  67   6      1          4        0   130
#> [334,]  69   0      0          0        1   148
#> [335,]  69   3      1          2        0   151
#> [336,]  65  14      1         13        1   150
#> [337,]  69   8      0        180        1   180
#> [338,]  71   7      0          7        0   230
#> [339,]  66   2      0          2        1   228
#> [340,]  71   6      0         45        1   158
#> [341,]  69   5      0          5        1   142
#> [342,]  71   3      0        103        0   133
#> [343,]  69   3      0          3        1   130
#> [344,]  67   1      0         36        1   104
#> [345,]  67   5      0          5        0   130
#> [346,]  69   8      1          5        1   195
#> [347,]  69   6      1          4        1   174
#> [348,]  72   3      1          0        1   132
#> [349,]  72   7      0          7        1   110
#> [350,]  69   8      1          7        1   108
#> [351,]  67   3      0        180        0   110
#> [352,]  69  19      0        180        0   130
#> [353,]  68  18      0         18        1   100
#> [354,]  67  14      0        172        1   140
#> [355,]  66   2      0        180        0   130
#> [356,]  67   7      1          4        0   122
#> [357,]  68   2      0          7        1   130
#> [358,]  70   3      0        123        0   130
#> [359,]  70   9      0        180        1   142
#> [360,]  72   5      1          4        0   170
#> [361,]  67   1      0          1        1    60
#> [362,]  67   4      0         60        1   136
#> [363,]  67   8      1          0        1   130
#> [364,]  68  10      1          8        1   160
#> [365,]  66  24      1         13        0   130
#> [366,]  70  35      1          0        1   105
#> [367,]  72  30      1          0        1   145
#> [368,]  70   7      0          7        0   102
#> [369,]  68   7      1          2        0   135
#> [370,]  71   6      0          9        0   120
#> [371,]  72  12      1         10        0   170
#> [372,]  67   8      0        180        1   170
#> [373,]  67   5      1          0        1   147
#> [374,]  67   9      0        180        0   158
#> [375,]  73  13      0        152        1   130
#> [376,]  72   6      1          5        0   115
#> [377,]  71   1      0        173        1   188
#> [378,]  70   3      0        180        0   121
#> [379,]  69   3      0        180        0   220
#> [380,]  71   3      1          2        0   150
#> [381,]  68   4      1          3        0   210
#> [382,]  72   5      0         28        0   120
#> [383,]  69  16      1         10        1   140
#> [384,]  69   8      1          1        0   164
#> [385,]  72  16      1          1        1   130
#> [386,]  70   4      0        180        0   180
#> [387,]  69   1      1          0        0   155
#> [388,]  73   7      0          7        1   140
#> [389,]  70   3      0          3        1   159
#> [390,]  73   0      0        180        1   161
#> [391,]  74   8      1          0        1    85
#> [392,]  69   2      1          0        1   110
#> [393,]  71   3      1          1        0   150
#> [394,]  68   9      0        180        1   120
#> [395,]  74   0      1          0        1    90
#> [396,]  73   3      1          0        1   136
#> [397,]  70   5      1          0        1   190
#> [398,]  71   8      1          7        0   149
#> [399,]  73  10      1          8        0   106
#> [400,]  69  12      1          1        1   149
#> [401,]  70  26      1         11        1   120
#> [402,]  74   4      0          4        0   120
#> [403,]  73   4      0         58        1   160
#> [404,]  70   3      0        180        1   154
#> [405,]  73   6      0        180        0   110
#> [406,]  71   7      1          2        0   143
#> [407,]  72   8      1          0        1   140
#> [408,]  74   3      0          3        1   150
#> [409,]  73  17      1         11        0   140
#> [410,]  71  13      1          8        0   121
#> [411,]  69   2      1          1        1    80
#> [412,]  71  14      1         13        1   170
#> [413,]  74   7      1          0        1   117
#> [414,]  72  10      1          8        1   153
#> [415,]  69   7      0        180        1   144
#> [416,]  70   8      0          8        0   120
#> [417,]  75   2      1          1        0   145
#> [418,]  72  10      1          9        1   160
#> [419,]  73  10      1         10        1   120
#> [420,]  74  15      1          9        1   179
#> [421,]  71   2      0         10        1   112
#> [422,]  73   1      0          1        1    80
#> [423,]  75  13      1          1        1   130
#> [424,]  72  15      1         12        1   120
#> [425,]  73  10      1          8        0   120
#> [426,]  72   1      1          1        0   168
#> [427,]  72   7      0         57        1   145
#> [428,]  73   5      1          3        1   112
#> [429,]  76  25      1         12        1   170
#> [430,]  75   1      0        180        1   140
#> [431,]  72   4      1          0        1   197
#> [432,]  73   5      0        180        0   126
#> [433,]  76   3      1          0        1   120
#> [434,]  72   5      0        180        0   154
#> [435,]  72   3      0        180        0   160
#> [436,]  75   3      1          1        0   180
#> [437,]  72   7      1          2        0   142
#> [438,]  73  15      0         15        1   160
#> [439,]  73  10      1         10        0   124
#> [440,]  74   7      0        180        1   150
#> [441,]  76   1      0        180        0   114
#> [442,]  74   2      1          1        0   140
#> [443,]  76   8      1          0        1   141
#> [444,]  74  19      1          4        1   200
#> [445,]  73   6      0          6        1   114
#> [446,]  75  23      1         14        1   110
#> [447,]  74   2      0        180        0   190
#> [448,]  72   4      0         85        1   120
#> [449,]  72   4      1          3        0   160
#> [450,]  75  12      0         12        1   160
#> [451,]  74   8      1          0        1   105
#> [452,]  76  13      1          8        1   148
#> [453,]  74   6      0        180        0   160
#> [454,]  76   4      0          4        1   155
#> [455,]  75   1      0          1        1   125
#> [456,]  73   1      0         52        1   105
#> [457,]  73   0      0        180        0   156
#> [458,]  72   5      0        180        0   120
#> [459,]  78  12      1         11        1   160
#> [460,]  76  44      1         10        0   105
#> [461,]  74   8      1          8        1   170
#> [462,]  73  33      1         12        1   175
#> [463,]  77  12      1          9        1   100
#> [464,]  73  10      1          9        0   146
#> [465,]  77  12      0        180        0   130
#> [466,]  77   1      1          0        1    90
#> [467,]  76  12      1         11        1   120
#> [468,]  78   5      1          0        1   170
#> [469,]  73   7      1          0        0   174
#> [470,]  74   6      0         79        1   140
#> [471,]  74   9      1          8        0   200
#> [472,]  75   6      0        180        0   150
#> [473,]  79  10      1          8        0   190
#> [474,]  74   2      1          0        1   130
#> [475,]  78  18      0         18        1   144
#> [476,]  76  29      0         47        0    90
#> [477,]  73   8      1          1        1   162
#> [478,]  74   2      0        180        0   100
#> [479,]  78   8      1          6        1   110
#> [480,]  76  13      1          1        1   170
#> [481,]  78  32      1          9        1   198
#> [482,]  79   6      0        180        0   170
#> [483,]  80  10      1          6        1   147
#> [484,]  78   0      0        180        1   212
#> [485,]  78  13      1          5        0   130
#> [486,]  75   5      0        119        1   150
#> [487,]  80   8      0          8        1   120
#> [488,]  74  10      1          8        0   135
#> [489,]  76   1      0          1        1    83
#> [490,]  79   4      0         80        0   145
#> [491,]  75   4      1          0        0   212
#> [492,]  77   2      1          0        1   143
#> [493,]  75  11      1          4        0   162
#> [494,]  75   3      0          3        0     0
#> [495,]  76   7      0         29        1   150
#> [496,]  77  24      0         24        1   160
#> [497,]  79   8      0         32        1   120
#> [498,]  80   9      0         23        1   128
#> [499,]  78   6      1          0        1   240
#> [500,]  78  11      1          1        1   140
#> [501,]  79   2      1          0        1   121
#> [502,]  76   4      0          4        1   160
#> [503,]  76  10      1          8        0   180
#> [504,]  77   6      0          6        1   107
#> [505,]  78  11      0        180        1   135
#> [506,]  76   1      0          1        1   140
#> [507,]  77  31      1          3        1   161
#> [508,]  78   7      1          0        1   110
#> [509,]  77   7      0        180        1   170
#> [510,]  79   4      1          0        1   120
#> [511,]  81   1      0        180        0   120
#> [512,]  80  15      1         12        1   150
#> [513,]  77   9      1          4        0   141
#> [514,]  82   5      0          8        1   120
#> [515,]  80  40      1          0        1   138
#> [516,]  78   4      0         59        1   112
#> [517,]  80  17      1         12        0   100
#> [518,]  76   7      0        161        0   151
#> [519,]  80  15      1          0        1    90
#> [520,]  81   4      1          2        1   126
#> [521,]  79  28      0        164        0   100
#> [522,]  79   1      0         37        1   140
#> [523,]  81   2      0        175        0   172
#> [524,]  77  13      1          0        1   190
#> [525,]  78  15      0         15        0   165
#> [526,]  78  26      1          5        0   194
#> [527,]  76   1      0        166        0   131
#> [528,]  81   4      1          1        1   104
#> [529,]  78  20      1          0        1   109
#> [530,]  80   1      0          1        0   100
#> [531,]  78   3      1          1        1   152
#> [532,]  77  10      1          8        1   130
#> [533,]  82   3      1          1        1   144
#> [534,]  77   5      0         85        0   188
#> [535,]  80   2      1          1        0   168
#> [536,]  79   6      0          6        0   152
#> [537,]  80   6      1          0        1   119
#> [538,]  80   5      0          5        1   130
#> [539,]  82   1      1          0        1    82
#> [540,]  79   1      0        125        0   193
#> [541,]  82  21      1          2        0   155
#> [542,]  79   4      0          4        1   121
#> [543,]  82   5      0        180        0   110
#> [544,]  83   5      0        180        0   148
#> [545,]  83   4      0        103        0    97
#> [546,]  81  11      1          8        0   160
#> [547,]  81   5      0        177        0    41
#> [548,]  80  11      1          8        0   170
#> [549,]  79   4      0          4        1   183
#> [550,]  78   9      1          4        1   120
#> [551,]  82   8      1          1        0   128
#> [552,]  79   1      0        180        1   170
#> [553,]  84   5      1          1        1    85
#> [554,]  81  20      1          9        0   170
#> [555,]  83   8      0          8        0   115
#> [556,]  80   6      1          0        1   150
#> [557,]  80  11      1          8        0   110
#> [558,]  81   8      0        180        0   146
#> [559,]  79   7      0        177        0   197
#> [560,]  79   0      1          0        1    96
#> [561,]  81   2      1          1        0   198
#> [562,]  83   2      0          2        1   155
#> [563,]  84   4      0        167        0   198
#> [564,]  84   4      0          4        1    85
#> [565,]  81   1      0          1        1   150
#> [566,]  81   4      0         90        1   138
#> [567,]  85   3      1          2        1   160
#> [568,]  80  13      1          8        1   140
#> [569,]  84   4      0         89        1   129
#> [570,]  79   4      0          4        1    60
#> [571,]  83   1      0          1        1   100
#> [572,]  82  19      0         19        0   120
#> [573,]  80  30      1         13        0   220
#> [574,]  83   9      0        180        0   198
#> [575,]  79  14      1          0        0   110
#> [576,]  83   3      0        114        0    98
#> [577,]  81  14      1         12        1   128
#> [578,]  83   2      0        154        0   130
#> [579,]  85   9      1          6        1   160
#> [580,]  83   1      0        180        0   160
#> [581,]  81   1      0          1        1   145
#> [582,]  81  12      0         12        1   163
#> [583,]  82   5      1          0        1   146
#> [584,]  81   4      0          4        0   160
#> [585,]  86  12      0        180        1   120
#> [586,]  83  12      1          2        1   170
#> [587,]  81  19      1         14        0   120
#> [588,]  82  15      1          0        0   183
#> [589,]  81  16      1          9        0   180
#> [590,]  84   6      0        165        0   145
#> [591,]  82   9      0        180        1   134
#> [592,]  84   3      0        180        1   120
#> [593,]  81  13      0        180        0   152
#> [594,]  85   3      0          3        1   118
#> [595,]  81   4      0        180        0   160
#> [596,]  82   1      0        180        1   193
#> [597,]  87   2      0          5        1   137
#> [598,]  86  12      1          0        1   132
#> [599,]  86   6      1          0        1   140
#> [600,]  84  16      0         70        1   150
#> [601,]  83  19      0         43        0   150
#> [602,]  84   3      1          2        0   125
#> [603,]  83  10      1          0        1   190
#> [604,]  84   7      1          2        0   148
#> [605,]  87   2      0        180        0   113
#> [606,]  84   9      0         92        1   110
#> [607,]  84   3      0        180        1   170
#> [608,]  82   4      0          4        0   130
#> [609,]  86  13      0        177        0   163
#> [610,]  86   6      0          6        1   117
#> [611,]  84  13      0         62        1   100
#> [612,]  86   6      1          1        0   112
#> [613,]  88   4      0          4        0   100
#> [614,]  83  20      1          3        1   150
#> [615,]  85  22      0         22        1   184
#> [616,]  83   9      0         65        1   150
#> [617,]  86   6      0         46        0   173
#> [618,]  88   3      0        115        0   110
#> [619,]  83   3      0          3        1   130
#> [620,]  87   8      0          8        1   157
#> [621,]  86  15      1          8        1   109
#> [622,]  88   4      0          4        0    86
#> [623,]  89   4      0          4        1   153
#> [624,]  89   5      0        119        1   140
#> [625,]  87   6      0        180        1   110
#> [626,]  87   1      0          1        0   170
#> [627,]  84   8      0        180        1   119
#> [628,]  85   8      0          8        1   136
#> [629,]  84   2      0        110        1   174
#> [630,]  84   0      0        180        1   136
#> [631,]  89  10      0         46        1   170
#> [632,]  88   1      0          1        0   135
#> [633,]  91   8      0          8        0   100
#> [634,]  87   2      0        180        0   160
#> [635,]  87   6      1          0        0   125
#> [636,]  86   3      1          0        1    80
#> [637,]  88   7      0         24        0   119
#> [638,]  88   8      0         50        1   154
#> [639,]  90  11      1         10        1   186
#> [640,]  87   6      0        126        1   168
#> [641,]  86   9      1          7        0   130
#> [642,]  91   1      0          1        1    74
#> [643,]  87   5      0         36        1   150
#> [644,]  90   5      1          0        1   125
#> [645,]  88   3      1          2        0   159
#> [646,]  89   3      1          1        1   160
#> [647,]  92   1      0          1        1   167
#> [648,]  91   3      0         33        1   137
#> [649,]  88   5      0        158        0   100
#> [650,]  87   7      0         74        1   105
#> [651,]  89   2      0        168        0   118
#> [652,]  89  52      0         52        1   130
#> [653,]  91   0      0          0        0     0
#> [654,]  89  14      0        180        1    84
#> [655,]  90  18      0        180        0   188
#> [656,]  91   4      1          0        1   120
#> [657,]  90  19      1         11        1   129
#> [658,]  94   6      0         50        0    78
#> [659,]  90   1      0          1        1   118
#> [660,]  91   2      0          2        1   116
#> [661,]  93   8      0        179        1   110
#> [662,]  94   8      0          8        1   142
#> [663,]  90   3      0         67        0   162
#> [664,]  95   8      1          5        1   150
#> [665,]  91  12      0         53        1   212
#> [666,]  93   0      1          0        1   122
#> [667,]  92   5      0         69        0   139
#> [668,]  92   2      0          2        0   112
#> [669,]  96   3      1          0        1   104
#> [670,]  96  15      1          0        1   140
#> 
#> $y
#>   [1] 180.0+   5.0+   5.0+ 180.0+ 180.0+ 115.0  180.0+ 180.0+  12.0    5.0+
#>  [11] 180.0+ 180.0+ 180.0+   2.0+ 180.0+   5.0+   2.0+ 180.0+ 180.0+   3.0 
#>  [21] 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 180.0+ 155.0+ 180.0+ 180.0+ 180.0+
#>  [31] 180.0+ 180.0+   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [41] 180.0+ 180.0+ 180.0+   6.0+ 180.0+ 180.0+  73.0    5.0+ 180.0+ 180.0+
#>  [51]   5.0+ 180.0+ 177.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 172.0+
#>  [61] 180.0+ 180.0+ 180.0+   7.0  180.0+ 180.0+   2.0    1.0  179.0+ 180.0+
#>  [71] 179.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [81] 180.0+ 180.0+   7.0   36.0   88.0+ 180.0+   4.0+ 180.0+ 180.0+ 180.0+
#>  [91] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  85.0  180.0+ 180.0+ 166.0+ 180.0+
#> [101] 152.0+   7.0+   6.0+ 180.0+ 180.0+ 171.0+ 174.0+  28.0    6.0+ 180.0+
#> [111]   9.0+ 180.0+ 180.0+   7.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  16.0+
#> [121] 180.0+  16.0   15.0+ 180.0+ 180.0+ 180.0+  12.0+ 134.0+ 180.0+   8.0 
#> [131]   2.0    3.0+ 180.0+ 180.0+ 180.0+ 180.0+ 165.0  180.0+ 180.0+ 180.0+
#> [141] 180.0+   8.0+ 180.0+ 180.0+   0.5  180.0+ 180.0+ 180.0+ 180.0+ 171.0+
#> [151]  15.0  166.0+   5.0+ 180.0+   4.0+ 147.0+ 180.0+ 180.0+   5.0+   4.0+
#> [161] 180.0+   9.0+ 180.0+ 180.0+   2.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [171] 180.0+   9.0+   0.5  180.0+ 180.0+   3.0    1.0  180.0+ 180.0+  10.0+
#> [181] 180.0+ 180.0+  17.0   45.0   19.0  180.0+ 180.0+   9.0+ 180.0+ 172.0+
#> [191]  24.0  180.0+   8.0  180.0+ 180.0+   1.0+  15.0  180.0+ 180.0+  77.0 
#> [201]   8.0+ 180.0+ 180.0+ 180.0+ 170.0  180.0+   7.0  180.0+ 180.0+ 180.0+
#> [211]   3.0+  30.0  180.0+  28.0   13.0+   5.0   18.0    5.0+ 180.0+ 180.0+
#> [221]   1.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   9.0   22.0 
#> [231] 180.0+ 180.0+  84.0    7.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0+   3.0+
#> [241] 167.0  180.0+  17.0  180.0+  36.0  180.0+   3.0+ 180.0+  12.0  180.0+
#> [251] 180.0+  12.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   9.0    3.0  180.0+
#> [261]  11.0+   2.0+  18.0+ 180.0+ 180.0+ 180.0+   3.0+   2.0+ 180.0+ 180.0+
#> [271]  13.0  179.0+ 166.0+  14.0+   3.0    0.5+   3.0+ 175.0+ 180.0+   7.0+
#> [281]  16.0  180.0+   1.0  180.0+ 180.0+ 180.0+ 123.0+  18.0   11.0+ 180.0+
#> [291]  79.0   80.0   15.0+ 180.0+   4.0+  15.0  174.0+ 180.0+ 180.0+ 180.0+
#> [301]   8.0+   3.0  180.0+  10.0  180.0+ 180.0+ 180.0+   6.0  180.0+  19.0+
#> [311]  12.0  179.0+  11.0+  18.0    7.0+ 180.0+ 180.0+ 152.0+ 180.0+  93.0 
#> [321]  21.0+ 180.0+   1.0  101.0    4.0  180.0+ 180.0+   1.0  180.0+ 180.0+
#> [331] 171.0  180.0+   6.0    0.5  180.0+  14.0+ 180.0+   7.0+   2.0   45.0 
#> [341]   5.0+ 103.0    3.0+  36.0    5.0+ 180.0+  97.0  180.0+   7.0    8.0+
#> [351] 180.0+ 180.0+  18.0  172.0+ 180.0+   7.0    7.0  123.0  180.0+ 180.0+
#> [361]   1.0   60.0  180.0+  10.0+ 180.0+ 180.0+ 162.0    7.0+   7.0+   9.0 
#> [371]  12.0  180.0+ 180.0+ 180.0+ 152.0  180.0+ 173.0+ 180.0+ 180.0+ 180.0+
#> [381] 180.0+  28.0   16.0+ 180.0+  16.0+ 180.0+ 180.0+   7.0+   3.0+ 180.0+
#> [391] 180.0+   2.0    3.0+ 180.0+   0.5  180.0+ 180.0+   8.0   87.0   12.0 
#> [401] 180.0+   4.0+  58.0  180.0+ 180.0+ 180.0+ 180.0+   3.0  180.0+ 175.0 
#> [411]   2.0   14.0+ 180.0+  10.0+ 180.0+   8.0+ 180.0+ 159.0   15.0  180.0+
#> [421]  10.0    1.0   13.0  180.0+  10.0    1.0   57.0    5.0  180.0+ 180.0+
#> [431] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   7.0   15.0+  10.0  180.0+
#> [441] 180.0+ 180.0+ 180.0+ 180.0+   6.0  180.0+ 180.0+  85.0  180.0+  12.0 
#> [451] 180.0+ 180.0+ 180.0+   4.0    1.0   52.0  180.0+ 180.0+  12.0  180.0+
#> [461]   8.0   33.0  180.0+ 180.0+ 180.0+   1.0   12.0  180.0+   7.0+  79.0 
#> [471] 168.0+ 180.0+ 180.0+ 176.0+  18.0   47.0  180.0+ 180.0+   8.0+ 180.0+
#> [481]  32.0  180.0+  10.0  180.0+ 172.0  119.0    8.0  180.0+   1.0   80.0 
#> [491]   4.0+   2.0  152.0+   3.0   29.0   24.0   32.0   23.0  180.0+ 180.0+
#> [501] 180.0+   4.0   10.0+   6.0  180.0+   1.0  171.0   43.0  180.0+ 138.0 
#> [511] 180.0+ 180.0+  71.0    8.0   40.0   59.0   17.0  161.0  180.0+  93.0 
#> [521] 164.0   37.0  175.0+  22.0   15.0+ 171.0+ 166.0+  71.0   20.0+   1.0 
#> [531]   3.0+  10.0  180.0+  85.0   10.0    6.0+   6.0    5.0    1.0  125.0 
#> [541] 180.0+   4.0  180.0+ 180.0+ 103.0  180.0+ 177.0+ 169.0    4.0  180.0+
#> [551] 180.0+ 180.0+ 180.0+  20.0    8.0+ 180.0+ 180.0+ 180.0+ 177.0+   0.5 
#> [561] 180.0+   2.0  167.0    4.0    1.0   90.0  180.0+ 180.0+  89.0    4.0 
#> [571]   1.0   19.0   30.0  180.0+ 180.0+ 114.0  180.0+ 154.0  180.0+ 180.0+
#> [581]   1.0   12.0    5.0+   4.0+ 180.0+  77.0  180.0+  83.0  180.0+ 165.0 
#> [591] 180.0+ 180.0+ 180.0+   3.0+ 180.0+ 180.0+   5.0  180.0+   6.0   70.0 
#> [601]  43.0  180.0+ 180.0+ 180.0+ 180.0+  92.0  180.0+   4.0  177.0    6.0+
#> [611]  62.0    6.0+   4.0+  20.0   22.0   65.0   46.0  115.0    3.0+   8.0+
#> [621] 180.0+   4.0    4.0  119.0  180.0+   1.0+ 180.0+   8.0  110.0  180.0+
#> [631]  46.0    1.0+   8.0  180.0+  25.0    3.0   24.0   50.0   11.0  126.0 
#> [641] 180.0+   1.0   36.0   89.0   75.0    3.0+   1.0   33.0  158.0   74.0 
#> [651] 168.0   52.0    0.5  180.0+ 180.0+   4.0  180.0+  50.0    1.0+   2.0 
#> [661] 179.0+   8.0+  67.0    8.0   53.0    0.5   69.0    2.0    3.0   15.0+
#> 
#> $weights
#> NULL
#> 
#> $offset
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Warning: Multiple lambdas have been fit. Lambda will be set to 0.01 (see parameter 's').

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8443068