GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.coxnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.coxnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Caution: This learner is different to learners calling glmnet::cv.glmnet() in that it does not use the internal optimization of parameter lambda. Instead, lambda needs to be tuned by the user (e.g., via mlr3tuning). When lambda is tuned, the glmnet will be trained for each tuning iteration. While fitting the whole path of lambdas would be more efficient, as is done by default in glmnet::glmnet(), tuning/selecting the parameter at prediction time (using parameter s) is currently not supported in mlr3 (at least not in efficient manner). Tuning the s parameter is, therefore, currently discouraged.

When the data are i.i.d. and efficiency is key, we recommend using the respective auto-tuning counterpart in mlr_learners_surv.cv_glmnet(). However, in some situations this is not applicable, usually when data are imbalanced or not i.i.d. (longitudinal, time-series) and tuning requires custom resampling strategies (blocked design, stratification).

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.coxnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exact	logical	FALSE	TRUE, FALSE	-

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	TRUE	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	0.01		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.logistic	character	Newton	Newton, modified.Newton	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvGlmnet

Methods

Public methods

LearnerSurvGlmnet$new()

LearnerSurvGlmnet$selected_features()

LearnerSurvGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("surv.glmnet")
print(learner)
#> 
#> ── <LearnerSurvGlmnet> (surv.glmnet): Regularized Generalized Linear Model ─────
#> • Model: -
#> • Parameters: use_pred_offset=TRUE
#> • Packages: mlr3, mlr3proba, mlr3extralearners, and glmnet
#> • Predict Types: [crank], distr, and lp
#> • Feature Types: logical, integer, and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: offset, selected_features, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("grace")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#>    Df  %Dev   Lambda
#> 1   0  0.00 0.200000
#> 2   1  0.68 0.182200
#> 3   2  1.48 0.166000
#> 4   2  2.45 0.151300
#> 5   2  3.28 0.137800
#> 6   3  6.01 0.125600
#> 7   3  8.67 0.114400
#> 8   3 10.65 0.104300
#> 9   3 12.19 0.095010
#> 10  3 13.43 0.086570
#> 11  3 14.44 0.078880
#> 12  3 15.28 0.071870
#> 13  3 15.97 0.065490
#> 14  3 16.55 0.059670
#> 15  4 17.06 0.054370
#> 16  4 17.49 0.049540
#> 17  4 17.87 0.045140
#> 18  5 18.21 0.041130
#> 19  6 18.50 0.037470
#> 20  6 18.76 0.034150
#> 21  6 18.98 0.031110
#> 22  6 19.17 0.028350
#> 23  6 19.33 0.025830
#> 24  6 19.46 0.023530
#> 25  6 19.58 0.021440
#> 26  6 19.67 0.019540
#> 27  6 19.75 0.017800
#> 28  6 19.82 0.016220
#> 29  6 19.88 0.014780
#> 30  6 19.93 0.013470
#> 31  6 19.97 0.012270
#> 32  6 20.00 0.011180
#> 33  6 20.03 0.010190
#> 34  6 20.06 0.009283
#> 35  6 20.08 0.008458
#> 36  6 20.09 0.007707
#> 37  6 20.11 0.007022
#> 38  6 20.12 0.006398
#> 39  6 20.13 0.005830
#> 40  6 20.14 0.005312
#> 41  6 20.15 0.004840
#> 42  6 20.15 0.004410
#> 43  6 20.16 0.004018
#> 44  6 20.16 0.003661
#> 45  6 20.17 0.003336
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  32   5      1          0        1   121
#>   [2,]  33   2      0          2        0   150
#>   [3,]  35   5      1          2        0   172
#>   [4,]  35  10      1          9        0   106
#>   [5,]  35   2      0        180        0   121
#>   [6,]  35   2      1          1        1   112
#>   [7,]  37   9      0        180        1   151
#>   [8,]  38  13      1          0        1   161
#>   [9,]  38   2      0        115        0   150
#>  [10,]  35   0      0        180        1   119
#>  [11,]  38  12      1          8        1   120
#>  [12,]  36   5      1          0        1   115
#>  [13,]  38  16      1         10        0   160
#>  [14,]  38  12      1         11        1    92
#>  [15,]  40  12      1          9        0   153
#>  [16,]  42   3      1          1        1   130
#>  [17,]  37   1      1          0        1   146
#>  [18,]  40   2      1          1        1   148
#>  [19,]  42   2      0        180        1   100
#>  [20,]  38   5      1          3        0   125
#>  [21,]  42   2      0          2        0   140
#>  [22,]  40  11      1         10        1   120
#>  [23,]  42   2      0        180        0   100
#>  [24,]  40   1      1          0        1   145
#>  [25,]  43   4      1          0        1   130
#>  [26,]  42   4      0        180        0   162
#>  [27,]  42  15      1         13        1   125
#>  [28,]  40   3      1          1        0   170
#>  [29,]  42   2      0        180        1   124
#>  [30,]  45   3      0        180        1   154
#>  [31,]  41  10      1          8        0   150
#>  [32,]  44   3      0        180        0   141
#>  [33,]  41  13      1          1        0   140
#>  [34,]  45   9      1          7        0   110
#>  [35,]  45   6      0        180        1   170
#>  [36,]  41   5      1          4        1   141
#>  [37,]  44   2      1          1        1   150
#>  [38,]  43   2      0        180        1   140
#>  [39,]  45   2      0        180        1   140
#>  [40,]  46   2      1          1        0   126
#>  [41,]  47   4      1          3        0   118
#>  [42,]  48  15      0        180        1   160
#>  [43,]  45   3      0        150        0   130
#>  [44,]  44   3      1          0        1   180
#>  [45,]  46   7      1          2        0   166
#>  [46,]  43  29      0        180        1   180
#>  [47,]  47   6      1          0        1   116
#>  [48,]  46  13      1         10        0   100
#>  [49,]  44   0      1          0        1    96
#>  [50,]  47   4      1          3        1   160
#>  [51,]  45   8      1          0        1   117
#>  [52,]  49   5      0         73        1   136
#>  [53,]  45   5      0          5        0   141
#>  [54,]  46   2      1          1        1   122
#>  [55,]  46   6      1          0        1   100
#>  [56,]  44   4      1          0        1   114
#>  [57,]  47   2      0        180        0   108
#>  [58,]  44   9      1          8        1   135
#>  [59,]  45   5      0        180        1   190
#>  [60,]  46   4      0        180        1   121
#>  [61,]  44   2      0        180        0   142
#>  [62,]  46  15      0        180        1   120
#>  [63,]  45   9      1          0        1   145
#>  [64,]  48   3      0        180        0   154
#>  [65,]  48  12      1         11        0   200
#>  [66,]  47   5      1          3        1   130
#>  [67,]  47   9      1          6        0   170
#>  [68,]  49   4      0        180        0   117
#>  [69,]  47  10      0         10        1   140
#>  [70,]  50   1      1          0        1   129
#>  [71,]  50   4      1          1        0   125
#>  [72,]  50   6      1          2        1   140
#>  [73,]  50   7      0        180        1   110
#>  [74,]  49  15      1         11        1   160
#>  [75,]  47   2      0        180        0   150
#>  [76,]  49  23      0        179        1   112
#>  [77,]  46   6      1          0        1   156
#>  [78,]  52   2      0        180        1   170
#>  [79,]  50   7      1          0        1    92
#>  [80,]  50   4      0          4        1   100
#>  [81,]  50   1      1          0        0   150
#>  [82,]  50   9      0        180        0   130
#>  [83,]  49   7      1          4        1    90
#>  [84,]  47   8      0        180        0   160
#>  [85,]  47   6      0        180        1   162
#>  [86,]  51   8      0        180        1   140
#>  [87,]  52   2      0        180        0   155
#>  [88,]  46   3      0        180        1   120
#>  [89,]  46   1      1          1        0   145
#>  [90,]  53   8      0         36        1   160
#>  [91,]  47   2      1          1        0   110
#>  [92,]  52   4      1          4        0   152
#>  [93,]  49   9      1          3        0   102
#>  [94,]  49  15      0        180        1   160
#>  [95,]  53   5      0        180        1   140
#>  [96,]  54  17      1         12        1   102
#>  [97,]  53   5      0         77        0   159
#>  [98,]  53   7      1          0        0   199
#>  [99,]  54   6      1          3        0   129
#> [100,]  50   2      0          5        1   106
#> [101,]  50  10      1          6        0   122
#> [102,]  50  14      1         13        0   170
#> [103,]  53   8      1          7        0   160
#> [104,]  51  25      1          1        0   202
#> [105,]  49   5      1          2        1   150
#> [106,]  48   6      0        180        0   160
#> [107,]  53   4      1          0        1   156
#> [108,]  51  13      0         99        1   160
#> [109,]  49  16      0         16        0   125
#> [110,]  55   3      1          1        0   150
#> [111,]  52   7      1          2        0   154
#> [112,]  55   4      1          2        0   150
#> [113,]  52   4      0        180        1   180
#> [114,]  54   4      1          0        1   121
#> [115,]  52   4      0        180        0   183
#> [116,]  50   3      0        174        1   153
#> [117,]  55  28      1         13        1   160
#> [118,]  49   6      1          0        1   130
#> [119,]  50   7      1          1        0   156
#> [120,]  53   8      1          0        1   130
#> [121,]  52   5      0        175        1   117
#> [122,]  55   1      0        180        0   127
#> [123,]  54   1      0        180        0   162
#> [124,]  54   7      1          0        1   100
#> [125,]  52   8      0        180        0   119
#> [126,]  53  18      1          9        1   150
#> [127,]  55   6      0        180        0   170
#> [128,]  53  10      1          9        0   172
#> [129,]  53  15      0         15        1    90
#> [130,]  55   6      0        180        1   100
#> [131,]  55   6      1          5        1   138
#> [132,]  54  12      1          0        1   190
#> [133,]  56   3      0          8        1   139
#> [134,]  55   1      0          2        0   130
#> [135,]  57   3      0          3        0   120
#> [136,]  52   9      1          3        0   170
#> [137,]  57   5      1          3        1   138
#> [138,]  57   1      0          1        1   100
#> [139,]  52   2      0        180        0   140
#> [140,]  55  11      1          7        0   104
#> [141,]  52  15      1         14        0   130
#> [142,]  56  14      1         11        0   130
#> [143,]  53   3      1          0        1   200
#> [144,]  57   0      0          0        1   150
#> [145,]  53  21      1         13        1   130
#> [146,]  58   6      1          0        1    90
#> [147,]  54  17      1          8        1   227
#> [148,]  55   9      1          2        1   147
#> [149,]  55  13      0        166        1   140
#> [150,]  53   4      0        147        1   145
#> [151,]  57  11      1         10        1   129
#> [152,]  55   3      1          2        0   140
#> [153,]  54   7      1          0        1   141
#> [154,]  56   4      0          4        0   164
#> [155,]  55   5      1          0        0   140
#> [156,]  56   7      1          5        1   120
#> [157,]  55   2      0          2        0   106
#> [158,]  59   9      1          1        1   125
#> [159,]  60  11      1          9        0   106
#> [160,]  59   3      0        180        0   120
#> [161,]  57   2      0          2        1   120
#> [162,]  60   5      1          1        0   138
#> [163,]  57   5      0        180        1   130
#> [164,]  58  11      1          9        1   124
#> [165,]  59   5      0        180        1   155
#> [166,]  58  26      1          0        1   189
#> [167,]  58   4      1          3        0   120
#> [168,]  59   2      1          1        0   140
#> [169,]  58   8      0        161        1   140
#> [170,]  58  14      1          6        0   190
#> [171,]  61   4      1          3        0   151
#> [172,]  61   9      1          8        0   150
#> [173,]  61   3      1          2        1   102
#> [174,]  61  20      1         13        0   130
#> [175,]  57   2      1          0        1   116
#> [176,]  58  10      0         10        1   150
#> [177,]  57   4      1          3        0   138
#> [178,]  57  11      0        180        1   150
#> [179,]  61   3      0         17        0   143
#> [180,]  56  14      0         45        0   130
#> [181,]  57   3      1          2        0   120
#> [182,]  58  19      1         13        1   140
#> [183,]  56  13      1          6        1   158
#> [184,]  56  18      1         11        1   165
#> [185,]  59   9      1          0        1    80
#> [186,]  55   4      1          3        1   160
#> [187,]  60  12      1          0        1   114
#> [188,]  56   8      1          8        0   120
#> [189,]  61  13      1         12        1   130
#> [190,]  57   1      0          1        0   126
#> [191,]  57  15      1         13        1   110
#> [192,]  59  10      0        180        0   160
#> [193,]  61   8      0         77        0   120
#> [194,]  61  13      0         13        0   210
#> [195,]  58   8      1          5        0   152
#> [196,]  62  10      1          0        1   153
#> [197,]  62   7      1          2        1   180
#> [198,]  61  28      1          7        0   133
#> [199,]  58   8      1          3        1   150
#> [200,]  57   7      0        169        0   180
#> [201,]  61   7      0          7        1   150
#> [202,]  61   6      0          6        0   134
#> [203,]  59  13      1          2        0   198
#> [204,]  62   4      1          0        0   160
#> [205,]  58   3      1          0        1   146
#> [206,]  62   4      1          3        0   173
#> [207,]  59  16      1          9        1   133
#> [208,]  63   6      0         28        1   120
#> [209,]  61   5      0          5        1   110
#> [210,]  61   5      0          5        1   160
#> [211,]  58  11      1          9        0   179
#> [212,]  63   1      0        180        1   130
#> [213,]  63   4      1          3        0   222
#> [214,]  62   3      0        180        1   105
#> [215,]  63   4      0        180        1   190
#> [216,]  63  15      1         10        1   126
#> [217,]  63   4      1          1        0   155
#> [218,]  60  18      1         13        0   132
#> [219,]  59   8      0        180        1   140
#> [220,]  61   9      1          9        1   150
#> [221,]  62   7      0          7        0   150
#> [222,]  59   4      0        180        0   196
#> [223,]  60   7      1          5        1   141
#> [224,]  59   5      1          1        0   148
#> [225,]  60   7      1          1        1    90
#> [226,]  63   1      0          1        0   162
#> [227,]  63   1      0          1        0   130
#> [228,]  62   6      0        180        0   170
#> [229,]  61  15      1         13        0   170
#> [230,]  59   4      0          4        0   149
#> [231,]  60   3      0          3        0   168
#> [232,]  64  10      1          9        0   160
#> [233,]  63  12      1         10        0   200
#> [234,]  59  10      0        180        1   130
#> [235,]  64  12      1         11        0   160
#> [236,]  63  10      1          0        1   148
#> [237,]  63  14      1          9        0   123
#> [238,]  65  36      1         11        0   140
#> [239,]  63   4      1          3        0   162
#> [240,]  66   3      1          1        0   127
#> [241,]  64  32      1          9        1   160
#> [242,]  63  12      1          9        0   114
#> [243,]  66   5      1          0        1   110
#> [244,]  64   0      0          0        1    90
#> [245,]  60   6      0        180        0   130
#> [246,]  61  12      1         11        0   154
#> [247,]  61   4      0        180        1   113
#> [248,]  65   3      0        180        1   190
#> [249,]  63  16      1          7        1   110
#> [250,]  66   6      1          1        1   130
#> [251,]  63  12      0         12        1   150
#> [252,]  62   3      1          1        1   199
#> [253,]  65   6      0          9        0   112
#> [254,]  65   3      1          0        1    80
#> [255,]  62  13      1         11        0   180
#> [256,]  67  11      0         11        1   100
#> [257,]  64   2      0          2        0   201
#> [258,]  66  18      1          5        0   142
#> [259,]  62   9      0        180        0   145
#> [260,]  61  14      1          5        0   140
#> [261,]  61  15      1         10        0   130
#> [262,]  63   3      1          2        0   120
#> [263,]  64  19      1          8        1   160
#> [264,]  67   6      0        180        1   170
#> [265,]  64   6      1          0        1   125
#> [266,]  66   7      1          0        1   115
#> [267,]  66  13      1          0        0   118
#> [268,]  65   3      0          3        0   105
#> [269,]  64   0      0          0        1   148
#> [270,]  66   6      1          0        1   140
#> [271,]  65   2      1          1        1   170
#> [272,]  68   1      0        180        1   166
#> [273,]  63   7      1          0        0   162
#> [274,]  67   8      1          1        1   130
#> [275,]  68   5      0          5        1    90
#> [276,]  68  18      0        180        1   260
#> [277,]  65  17      1         14        1   100
#> [278,]  63   8      1          1        1   162
#> [279,]  65  18      1          3        0   120
#> [280,]  63   1      1          0        1   155
#> [281,]  63  10      0         18        1   130
#> [282,]  67  11      0         11        0   150
#> [283,]  66  12      1         10        1   150
#> [284,]  66  11      1          0        0   100
#> [285,]  65   4      1          2        1   145
#> [286,]  69  12      0         15        1   140
#> [287,]  66  15      1         13        1   160
#> [288,]  69  21      1         10        0   180
#> [289,]  69   6      0        180        1   100
#> [290,]  66   9      1          8        0   130
#> [291,]  63   8      0        180        1   120
#> [292,]  68  14      1         13        1   140
#> [293,]  66   3      0          3        1   138
#> [294,]  69   1      1          0        0   170
#> [295,]  67   1      0        180        1   160
#> [296,]  68  10      1         10        1   150
#> [297,]  67   7      1          4        1   130
#> [298,]  65   6      0          6        0    80
#> [299,]  67  12      1         12        0   160
#> [300,]  69   6      0         99        1   140
#> [301,]  64   4      0        179        0   160
#> [302,]  66   4      0        180        1   130
#> [303,]  64  11      0         11        0   125
#> [304,]  64   4      0        180        1   140
#> [305,]  67   2      0         18        0   131
#> [306,]  66   7      1          5        1   131
#> [307,]  66   4      0        180        0   177
#> [308,]  68   4      1          0        1   160
#> [309,]  69   4      1          3        1   150
#> [310,]  65  13      1         12        1   130
#> [311,]  64  21      0         21        1   155
#> [312,]  65   1      0          1        1   120
#> [313,]  68  18      1          0        1   160
#> [314,]  65   6      0        101        1   115
#> [315,]  68   4      0          4        1   190
#> [316,]  68   7      0        150        0   210
#> [317,]  67   2      0        180        0   128
#> [318,]  66   9      1          3        1   151
#> [319,]  70   4      1          0        1   180
#> [320,]  69   8      0        180        1   153
#> [321,]  66   4      0        180        0   130
#> [322,]  67  10      1          9        0   200
#> [323,]  67   6      1          4        0   130
#> [324,]  68  18      1         14        1   170
#> [325,]  65   2      0        180        0   130
#> [326,]  68   7      1          0        1   150
#> [327,]  69   3      1          2        0   151
#> [328,]  67  14      1         13        0   130
#> [329,]  65  14      1         13        1   150
#> [330,]  71   7      0          7        0   230
#> [331,]  66   2      0          2        1   228
#> [332,]  69   5      0          5        1   142
#> [333,]  71   3      0        103        0   133
#> [334,]  69   3      0          3        1   130
#> [335,]  70  22      1         13        0   103
#> [336,]  67   1      0         36        1   104
#> [337,]  68   6      0        180        0   145
#> [338,]  69   8      1          5        1   195
#> [339,]  69   6      1          4        1   174
#> [340,]  72   3      1          0        1   132
#> [341,]  72   7      0          7        1   110
#> [342,]  69   8      1          7        1   108
#> [343,]  67   3      0        180        0   110
#> [344,]  66   2      1          1        0   123
#> [345,]  69  19      0        180        0   130
#> [346,]  67  14      0        172        1   140
#> [347,]  66   2      0        180        0   130
#> [348,]  68   2      0          7        1   130
#> [349,]  67  13      1          9        0   130
#> [350,]  67  22      1          1        1   140
#> [351,]  68   3      0         19        0   135
#> [352,]  67  12      1          8        0   120
#> [353,]  67   1      0          1        1    60
#> [354,]  67   4      0         60        1   136
#> [355,]  69   5      0         76        0   120
#> [356,]  67   8      1          0        1   130
#> [357,]  72  13      1         11        1   195
#> [358,]  68  10      1          8        1   160
#> [359,]  66  24      1         13        0   130
#> [360,]  70  35      1          0        1   105
#> [361,]  68   7      1          2        0   135
#> [362,]  73  20      1          0        1   170
#> [363,]  72  12      1         10        0   170
#> [364,]  67   8      0        180        1   170
#> [365,]  67   5      1          0        1   147
#> [366,]  73  13      0        152        1   130
#> [367,]  70   5      0        180        0   150
#> [368,]  72   2      0          2        1   100
#> [369,]  67   4      1          1        0   134
#> [370,]  72   6      1          5        0   115
#> [371,]  71   1      0        173        1   188
#> [372,]  68  23      0        180        1   220
#> [373,]  70   3      0        180        0   121
#> [374,]  69   3      0        180        0   220
#> [375,]  68   4      1          3        0   210
#> [376,]  72   5      0         28        0   120
#> [377,]  71   5      0        180        0   191
#> [378,]  73   6      0        180        1   117
#> [379,]  69  16      1         10        1   140
#> [380,]  69   8      1          1        0   164
#> [381,]  68   7      0        180        1   130
#> [382,]  72  16      1          1        1   130
#> [383,]  70   4      0        180        0   180
#> [384,]  69   1      1          0        0   155
#> [385,]  73   6      1          0        1   270
#> [386,]  71   2      1          0        1   180
#> [387,]  73   7      0          7        1   140
#> [388,]  68  15      1         13        1   130
#> [389,]  70  13      1          9        0   100
#> [390,]  73   0      0        180        1   161
#> [391,]  74   8      1          0        1    85
#> [392,]  69   2      1          0        1   110
#> [393,]  68   9      0        180        1   120
#> [394,]  74   0      1          0        1    90
#> [395,]  70   5      1          0        1   190
#> [396,]  71  17      1         11        0   160
#> [397,]  71   3      1          2        1   190
#> [398,]  73  10      1          8        0   106
#> [399,]  69  12      1          1        1   149
#> [400,]  74   4      0          4        0   120
#> [401,]  72   5      1          3        1   160
#> [402,]  70   3      0        180        1   154
#> [403,]  73   6      0        180        0   110
#> [404,]  72  15      1          0        1   150
#> [405,]  73  17      1         11        0   140
#> [406,]  69   2      1          1        1    80
#> [407,]  70   4      1          0        1   140
#> [408,]  74   7      1          0        1   117
#> [409,]  72  10      1          8        1   153
#> [410,]  72  15      1         13        0   156
#> [411,]  71  10      1          9        1   120
#> [412,]  75   1      0          1        0   133
#> [413,]  73  10      1          9        1   146
#> [414,]  72  10      1          9        1   160
#> [415,]  73  10      1         10        1   120
#> [416,]  74  15      1          9        1   179
#> [417,]  71   2      0         10        1   112
#> [418,]  73   1      0          1        1    80
#> [419,]  75  13      1          1        1   130
#> [420,]  72  15      1         12        1   120
#> [421,]  70   7      1          4        0   184
#> [422,]  73  10      0        180        0   162
#> [423,]  72  11      0         11        1   140
#> [424,]  70   3      0          3        0   150
#> [425,]  73   5      1          3        1   112
#> [426,]  76  25      1         12        1   170
#> [427,]  75   1      0        180        1   140
#> [428,]  72   4      1          0        1   197
#> [429,]  71   3      1          0        0   144
#> [430,]  73   4      0        180        0   124
#> [431,]  74  34      1          8        1   233
#> [432,]  76   3      1          0        1   120
#> [433,]  71  32      1         12        1   107
#> [434,]  72   5      0        180        0   154
#> [435,]  72   3      0        180        0   160
#> [436,]  76   5      0          5        1   130
#> [437,]  77  11      0         11        1   150
#> [438,]  77   4      0          4        0   185
#> [439,]  72   7      1          2        0   142
#> [440,]  71  16      0        180        0   140
#> [441,]  74   3      0          3        1   128
#> [442,]  76   1      0        180        0   114
#> [443,]  74   2      1          1        0   140
#> [444,]  76   8      1          0        1   141
#> [445,]  74  19      1          4        1   200
#> [446,]  73   6      0          6        1   114
#> [447,]  75  23      1         14        1   110
#> [448,]  72   4      0         85        1   120
#> [449,]  72   4      1          3        0   160
#> [450,]  76  17      1          0        1   200
#> [451,]  73   4      1          3        1   125
#> [452,]  76  13      1         10        0   110
#> [453,]  75   7      0          7        0   190
#> [454,]  75   0      0          0        1   130
#> [455,]  73  13      1         11        0   195
#> [456,]  74   8      1          0        1   105
#> [457,]  76  13      1          8        1   148
#> [458,]  75   4      1          2        1   188
#> [459,]  76   4      0          4        1   155
#> [460,]  75   1      0          1        1   125
#> [461,]  73   1      0         52        1   105
#> [462,]  73   0      0        180        0   156
#> [463,]  72   5      0        180        0   120
#> [464,]  74  10      1          0        1   135
#> [465,]  76   5      1          0        1   167
#> [466,]  74   8      1          8        1   170
#> [467,]  75   9      0        180        1   140
#> [468,]  73  33      1         12        1   175
#> [469,]  77   5      1          0        0   123
#> [470,]  73  10      1          9        0   146
#> [471,]  78   5      1          0        1   170
#> [472,]  75   3      1          1        1   171
#> [473,]  75   6      0        180        0   150
#> [474,]  74   2      1          0        1   130
#> [475,]  78  18      0         18        1   144
#> [476,]  73   8      1          1        1   162
#> [477,]  73  11      1          2        1   110
#> [478,]  78   7      0          7        1   133
#> [479,]  74  15      0        180        1   172
#> [480,]  78   8      1          6        1   110
#> [481,]  76  13      1          1        1   170
#> [482,]  78  32      1          9        1   198
#> [483,]  79   6      0        180        0   170
#> [484,]  78   0      0        180        1   212
#> [485,]  78  13      1          5        0   130
#> [486,]  75   5      0        119        1   150
#> [487,]  75  12      1          1        1   120
#> [488,]  79   4      0         80        0   145
#> [489,]  78  12      1          9        0   150
#> [490,]  75   4      1          0        0   212
#> [491,]  77   2      1          0        1   143
#> [492,]  76  11      1          0        0   120
#> [493,]  75  11      1          4        0   162
#> [494,]  75   3      0          3        0     0
#> [495,]  80   9      0         23        1   128
#> [496,]  80   6      0          6        1   150
#> [497,]  78   6      1          0        1   240
#> [498,]  79  11      0        180        0   160
#> [499,]  79   2      1          0        1   121
#> [500,]  76   4      0          4        1   160
#> [501,]  79   4      0          4        1   125
#> [502,]  77   6      0          6        1   107
#> [503,]  80   3      1          0        1   120
#> [504,]  75   2      1          1        1   204
#> [505,]  78  11      0        180        1   135
#> [506,]  76   1      0          1        1   140
#> [507,]  77  31      1          3        1   161
#> [508,]  76   1      0          1        1    90
#> [509,]  78   7      1          0        1   110
#> [510,]  79   3      0          3        0   120
#> [511,]  77   7      0        180        1   170
#> [512,]  77   6      0          6        1   144
#> [513,]  79   4      1          0        1   120
#> [514,]  81   1      0        180        0   120
#> [515,]  80  15      1         12        1   150
#> [516,]  77   9      1          4        0   141
#> [517,]  78   4      0         59        1   112
#> [518,]  80  17      1         12        0   100
#> [519,]  76   7      0        161        0   151
#> [520,]  79  10      0         10        1   120
#> [521,]  80  15      1          0        1    90
#> [522,]  79  28      0        164        0   100
#> [523,]  80   9      0        118        1   186
#> [524,]  78  32      0        180        1   130
#> [525,]  79   1      0         37        1   140
#> [526,]  81   3      0        180        0   184
#> [527,]  81   2      0        175        0   172
#> [528,]  77  13      1          0        1   190
#> [529,]  80   5      1          1        1   108
#> [530,]  78   4      0        180        0   175
#> [531,]  79   3      0          3        1   101
#> [532,]  76   1      0        166        0   131
#> [533,]  81   4      1          1        1   104
#> [534,]  78  20      1          0        1   109
#> [535,]  80   1      0          1        0   100
#> [536,]  78   3      1          1        1   152
#> [537,]  77   5      0         85        0   188
#> [538,]  80   2      1          1        0   168
#> [539,]  79   6      0          6        0   152
#> [540,]  80   6      1          0        1   119
#> [541,]  78   2      0        180        0   148
#> [542,]  80   5      0          5        1   130
#> [543,]  82   1      1          0        1    82
#> [544,]  77   4      0        180        1    98
#> [545,]  78  12      0        180        0   134
#> [546,]  82  21      1          2        0   155
#> [547,]  84  22      1         10        0   180
#> [548,]  79   4      0          4        1   121
#> [549,]  80   6      0          6        1   110
#> [550,]  83   9      1          5        1   170
#> [551,]  82   5      0        180        0   110
#> [552,]  83   5      0        180        0   148
#> [553,]  79   7      1          6        0   130
#> [554,]  81  11      1          8        0   160
#> [555,]  81   5      0        177        0    41
#> [556,]  80  11      1          8        0   170
#> [557,]  78  23      1         10        1   145
#> [558,]  78   9      1          4        1   120
#> [559,]  81  15      0        180        1   140
#> [560,]  81  16      0         16        1   110
#> [561,]  81   8      0        180        0   146
#> [562,]  79   0      1          0        1    96
#> [563,]  85   4      0        180        0    90
#> [564,]  81   2      1          1        0   198
#> [565,]  82   6      0        128        1   100
#> [566,]  84   4      0        167        0   198
#> [567,]  80   3      1          1        1   230
#> [568,]  81   1      0          1        1   150
#> [569,]  84   1      0         38        1   205
#> [570,]  83   3      0        180        0   174
#> [571,]  81   4      0         90        1   138
#> [572,]  84   4      0         89        1   129
#> [573,]  80   6      0         71        1   189
#> [574,]  82  19      0         19        0   120
#> [575,]  79  14      1          0        0   110
#> [576,]  83   2      0        154        0   130
#> [577,]  82   0      0          2        1   100
#> [578,]  85   9      1          6        1   160
#> [579,]  81   4      0          4        0   175
#> [580,]  81  12      0         12        1   163
#> [581,]  82  16      1          8        0   103
#> [582,]  81   4      0          4        0   160
#> [583,]  86  12      0        180        1   120
#> [584,]  83  12      1          2        1   170
#> [585,]  81  19      1         14        0   120
#> [586,]  82  15      1          0        0   183
#> [587,]  80   2      0         88        0   135
#> [588,]  86   8      0          8        1   132
#> [589,]  86   3      0          3        1   140
#> [590,]  82   9      0        180        1   134
#> [591,]  84   3      0        180        1   120
#> [592,]  81  13      0        180        0   152
#> [593,]  85   3      0          3        1   118
#> [594,]  81   2      1          0        1   118
#> [595,]  81   4      0        180        0   160
#> [596,]  83   9      0        180        1   149
#> [597,]  87   2      0          5        1   137
#> [598,]  86  12      1          0        1   132
#> [599,]  82  14      1         11        1   103
#> [600,]  86   6      1          0        1   140
#> [601,]  84  16      0         70        1   150
#> [602,]  84   3      1          2        0   125
#> [603,]  83  10      1          0        1   190
#> [604,]  88  14      1          3        1   130
#> [605,]  83  13      1         12        0   170
#> [606,]  84   7      1          2        0   148
#> [607,]  86   4      0         38        1   122
#> [608,]  82   4      0          4        0   130
#> [609,]  86  13      0        177        0   163
#> [610,]  86   6      0          6        1   117
#> [611,]  84  13      0         62        1   100
#> [612,]  86   6      1          1        0   112
#> [613,]  88   4      0          4        0   100
#> [614,]  83  20      1          3        1   150
#> [615,]  88   4      0          4        1   115
#> [616,]  85  22      0         22        1   184
#> [617,]  83   9      0         65        1   150
#> [618,]  86   9      1          7        1   142
#> [619,]  87   2      0        180        1   130
#> [620,]  88   3      0        115        0   110
#> [621,]  88   2      0        180        1    68
#> [622,]  87   8      0          8        1   157
#> [623,]  86  15      1          8        1   109
#> [624,]  87   6      0        180        1   110
#> [625,]  87   1      0          1        0   170
#> [626,]  84   8      0        180        1   119
#> [627,]  85   8      0          8        1   136
#> [628,]  84   2      0        110        1   174
#> [629,]  87  29      0         29        1    97
#> [630,]  87  15      1          9        1   138
#> [631,]  84   0      0        180        1   136
#> [632,]  89  10      0         46        1   170
#> [633,]  90  14      0         14        1   100
#> [634,]  88   1      0          1        0   135
#> [635,]  91   8      0          8        0   100
#> [636,]  87   2      0        180        0   160
#> [637,]  87   6      1          0        0   125
#> [638,]  86   3      1          0        1    80
#> [639,]  88   7      0         24        0   119
#> [640,]  90  11      1         10        1   186
#> [641,]  86  10      0        180        1   137
#> [642,]  86   9      1          7        0   130
#> [643,]  90   4      1          0        0   121
#> [644,]  91   1      0          1        1    74
#> [645,]  87  43      0        178        1   130
#> [646,]  90   5      1          0        1   125
#> [647,]  92   1      0          1        1   167
#> [648,]  91   3      0         33        1   137
#> [649,]  88   5      0        158        0   100
#> [650,]  89  12      1          0        1   130
#> [651,]  89   2      0        168        0   118
#> [652,]  89  52      0         52        1   130
#> [653,]  92   7      0          7        1   110
#> [654,]  89  14      0        180        1    84
#> [655,]  90  18      0        180        0   188
#> [656,]  90  19      1         11        1   129
#> [657,]  94   6      0         50        0    78
#> [658,]  93   8      0        179        1   110
#> [659,]  92   4      0         76        1   149
#> [660,]  91   1      0        180        0   158
#> [661,]  90  16      0         16        1   106
#> [662,]  90   3      0         67        0   162
#> [663,]  96   3      0         12        1    97
#> [664,]  95   8      1          5        1   150
#> [665,]  91  12      0         53        1   212
#> [666,]  91   7      0          7        0   135
#> [667,]  92   5      0         69        0   139
#> [668,]  92   2      0          2        0   112
#> [669,]  93   4      0        180        1   135
#> [670,]  96   3      1          0        1   104
#> 
#> $y
#>   [1]   5.0+   2.0+   5.0+ 180.0+ 180.0+   2.0+ 180.0+ 180.0+ 115.0  180.0+
#>  [11]  12.0    5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 180.0+   5.0+
#>  [21]   2.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [31] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [41] 180.0+ 180.0+ 150.0  180.0+ 180.0+ 180.0+   6.0+ 180.0+ 180.0+ 180.0+
#>  [51] 180.0+  73.0    5.0+ 161.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [61] 180.0+ 180.0+ 177.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  10.0+ 172.0+
#>  [71] 180.0+ 180.0+ 180.0+ 179.0+ 180.0+ 179.0+ 180.0+ 180.0+ 180.0+   4.0+
#>  [81] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  36.0 
#>  [91] 180.0+   4.0+ 180.0+ 180.0+ 180.0+ 180.0+  77.0  180.0+ 180.0+   5.0 
#> [101] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 166.0+  99.0   16.0+ 180.0+
#> [111]   7.0+ 180.0+ 180.0+ 180.0+ 180.0+ 174.0+  28.0    6.0+ 180.0+ 180.0+
#> [121] 175.0+ 180.0+ 180.0+   7.0+ 180.0+ 180.0+ 180.0+ 180.0+  15.0+ 180.0+
#> [131] 180.0+  12.0+   8.0    2.0    3.0+ 180.0+ 140.0    1.0  180.0+ 180.0+
#> [141] 180.0+ 180.0+ 180.0+   0.5  180.0+ 180.0+ 171.0+  15.0  166.0+ 147.0+
#> [151] 180.0+ 180.0+ 180.0+   4.0+ 180.0+ 180.0+   2.0+ 180.0+ 180.0+ 180.0+
#> [161]   2.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 161.0+ 171.0+
#> [171] 180.0+ 180.0+   3.0  180.0+ 180.0+  10.0+ 180.0+ 180.0+  17.0   45.0 
#> [181]   3.0+  19.0  180.0+ 180.0+   9.0+ 180.0+ 172.0+   8.0  180.0+   1.0+
#> [191]  15.0  180.0+  77.0   13.0+   8.0+ 180.0+ 180.0+  94.0  180.0+ 169.0 
#> [201]   7.0    6.0  180.0+ 180.0+   3.0+ 180.0+ 180.0+  28.0    5.0    5.0+
#> [211] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+ 180.0+ 180.0+
#> [221]   7.0+ 180.0+  84.0  180.0+ 180.0+   1.0    1.0  180.0+ 180.0+   4.0+
#> [231]   3.0+ 167.0  180.0+ 180.0+  12.0  180.0+  14.0+  36.0  180.0+   3.0+
#> [241] 180.0+  12.0  180.0+   0.5  180.0+  12.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [251]  12.0  180.0+   9.0    3.0  180.0+  11.0+   2.0+  18.0+ 180.0+ 180.0+
#> [261] 180.0+   3.0+ 103.0  180.0+ 180.0+ 179.0+ 166.0+   3.0    0.5+ 180.0+
#> [271] 175.0+ 180.0+   7.0+   8.0    5.0  180.0+ 180.0+ 180.0+ 123.0+   1.0+
#> [281]  18.0   11.0+  80.0  180.0+   4.0+  15.0  180.0+ 174.0+ 180.0+ 180.0+
#> [291] 180.0+ 180.0+   3.0  175.0  180.0+  10.0  180.0+   6.0   12.0   99.0 
#> [301] 179.0+ 180.0+  11.0+ 180.0+  18.0    7.0+ 180.0+ 180.0+ 152.0+ 180.0+
#> [311]  21.0+   1.0   18.0+ 101.0    4.0  150.0  180.0+ 180.0+ 180.0+ 180.0+
#> [321] 180.0+ 174.0+   6.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+  14.0+   7.0+
#> [331]   2.0    5.0+ 103.0    3.0+ 180.0+  36.0  180.0+ 180.0+  97.0  180.0+
#> [341]   7.0    8.0+ 180.0+   2.0+ 180.0+ 172.0+ 180.0+   7.0   13.0+  51.0 
#> [351]  19.0  180.0+   1.0   60.0   76.0  180.0+ 132.0   10.0+ 180.0+ 180.0+
#> [361]   7.0+ 124.0   12.0  180.0+ 180.0+ 152.0  180.0+   2.0   76.0  180.0+
#> [371] 173.0+ 180.0+ 180.0+ 180.0+ 180.0+  28.0  180.0+ 180.0+  16.0+ 180.0+
#> [381] 180.0+  16.0+ 180.0+ 180.0+   6.0  180.0+   7.0+  15.0   13.0+ 180.0+
#> [391] 180.0+   2.0  180.0+   0.5  180.0+ 180.0+   3.0   87.0   12.0    4.0+
#> [401] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0  180.0+ 180.0+  10.0+ 180.0+
#> [411] 179.0+   1.0  180.0+ 159.0   15.0  180.0+  10.0    1.0   13.0  180.0+
#> [421] 104.0+ 180.0+  11.0    3.0+   5.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [431]  34.0  180.0+ 177.0+ 180.0+ 180.0+   5.0   11.0+   4.0+   7.0  180.0+
#> [441]   3.0  180.0+ 180.0+ 180.0+ 180.0+   6.0  180.0+  85.0  180.0+  17.0+
#> [451] 180.0+ 174.0+   7.0    0.5  180.0+ 180.0+ 180.0+  46.0    4.0    1.0 
#> [461]  52.0  180.0+ 180.0+ 180.0+ 180.0+   8.0  180.0+  33.0    5.0  180.0+
#> [471] 180.0+   3.0  180.0+ 176.0+  18.0  180.0+  11.0    7.0  180.0+   8.0+
#> [481] 180.0+  32.0  180.0+ 180.0+ 172.0  119.0   12.0   80.0  180.0+   4.0+
#> [491]   2.0   11.0  152.0+   3.0   23.0    6.0  180.0+ 180.0+ 180.0+   4.0 
#> [501]   4.0    6.0    3.0+   2.0+ 180.0+   1.0  171.0    1.0   43.0    3.0 
#> [511] 180.0+   6.0  138.0  180.0+ 180.0+  71.0   59.0   17.0  161.0   10.0+
#> [521] 180.0+ 164.0  118.0  180.0+  37.0  180.0+ 175.0+  22.0    5.0+ 180.0+
#> [531]   3.0  166.0+  71.0   20.0+   1.0    3.0+  85.0   10.0    6.0+   6.0 
#> [541] 180.0+   5.0    1.0  180.0+ 180.0+ 180.0+ 180.0+   4.0    6.0    9.0+
#> [551] 180.0+ 180.0+ 180.0+ 180.0+ 177.0+ 169.0   70.0  180.0+ 180.0+  16.0 
#> [561] 180.0+   0.5  180.0+ 180.0+ 128.0  167.0    3.0+   1.0   38.0  180.0+
#> [571]  90.0   89.0   71.0   19.0  180.0+ 154.0    2.0  180.0+   4.0+  12.0 
#> [581]  16.0+   4.0+ 180.0+  77.0  180.0+  83.0   88.0    8.0    3.0  180.0+
#> [591] 180.0+ 180.0+   3.0+ 180.0+ 180.0+ 180.0+   5.0  180.0+ 174.0    6.0 
#> [601]  70.0  180.0+ 180.0+  14.0   13.0  180.0+  38.0    4.0  177.0    6.0+
#> [611]  62.0    6.0+   4.0+  20.0    4.0   22.0   65.0   11.0  180.0+ 115.0 
#> [621] 180.0+   8.0+ 180.0+ 180.0+   1.0+ 180.0+   8.0  110.0   29.0  180.0+
#> [631] 180.0+  46.0   14.0    1.0+   8.0  180.0+  25.0    3.0   24.0   11.0 
#> [641] 180.0+ 180.0+   4.0    1.0  178.0+  89.0    1.0   33.0  158.0  180.0+
#> [651] 168.0   52.0    7.0  180.0+ 180.0+ 180.0+  50.0  179.0+  76.0  180.0+
#> [661]  16.0   67.0   12.0    8.0   53.0    7.0+  69.0    2.0  180.0+   3.0 
#> 
#> $weights
#> NULL
#> 
#> $offset
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Warning: Multiple lambdas have been fit. Lambda will be set to 0.01 (see parameter 's').

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8097697