GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.coxnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.coxnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Caution: This learner is different to learners calling glmnet::cv.glmnet() in that it does not use the internal optimization of parameter lambda. Instead, lambda needs to be tuned by the user (e.g., via mlr3tuning). When lambda is tuned, the glmnet will be trained for each tuning iteration. While fitting the whole path of lambdas would be more efficient, as is done by default in glmnet::glmnet(), tuning/selecting the parameter at prediction time (using parameter s) is currently not supported in mlr3 (at least not in efficient manner). Tuning the s parameter is, therefore, currently discouraged.

When the data are i.i.d. and efficiency is key, we recommend using the respective auto-tuning counterpart in mlr_learners_surv.cv_glmnet(). However, in some situations this is not applicable, usually when data are imbalanced or not i.i.d. (longitudinal, time-series) and tuning requires custom resampling strategies (blocked design, stratification).

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.coxnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exact	logical	FALSE	TRUE, FALSE	-

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	TRUE	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	0.01		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.logistic	character	Newton	Newton, modified.Newton	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvGlmnet

Methods

Public methods

LearnerSurvGlmnet$new()

LearnerSurvGlmnet$selected_features()

LearnerSurvGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.glmnet")
print(learner)
#> 
#> ── <LearnerSurvGlmnet> (surv.glmnet): Regularized Generalized Linear Model ─────
#> • Model: -
#> • Parameters: use_pred_offset=TRUE
#> • Packages: mlr3, mlr3proba, mlr3extralearners, and glmnet
#> • Predict Types: [crank], distr, and lp
#> • Feature Types: logical, integer, and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: offset, selected_features, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#>    Df  %Dev   Lambda
#> 1   0  0.00 0.164600
#> 2   2  0.85 0.150000
#> 3   2  1.68 0.136700
#> 4   3  4.38 0.124500
#> 5   3  6.85 0.113500
#> 6   3  8.71 0.103400
#> 7   3 10.16 0.094190
#> 8   3 11.32 0.085830
#> 9   3 12.27 0.078200
#> 10  3 13.05 0.071250
#> 11  3 13.70 0.064920
#> 12  3 14.24 0.059160
#> 13  3 14.70 0.053900
#> 14  4 15.09 0.049110
#> 15  4 15.43 0.044750
#> 16  5 15.73 0.040770
#> 17  5 15.98 0.037150
#> 18  5 16.20 0.033850
#> 19  5 16.39 0.030840
#> 20  5 16.55 0.028100
#> 21  5 16.68 0.025610
#> 22  6 16.81 0.023330
#> 23  6 16.92 0.021260
#> 24  6 17.01 0.019370
#> 25  6 17.08 0.017650
#> 26  6 17.15 0.016080
#> 27  6 17.20 0.014650
#> 28  6 17.25 0.013350
#> 29  6 17.29 0.012170
#> 30  6 17.32 0.011080
#> 31  6 17.35 0.010100
#> 32  6 17.37 0.009203
#> 33  6 17.39 0.008385
#> 34  6 17.40 0.007640
#> 35  6 17.42 0.006962
#> 36  6 17.43 0.006343
#> 37  6 17.44 0.005780
#> 38  6 17.45 0.005266
#> 39  6 17.45 0.004798
#> 40  6 17.46 0.004372
#> 41  6 17.46 0.003984
#> 42  6 17.47 0.003630
#> 43  6 17.47 0.003307
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  28   9      0        180        1   107
#>   [2,]  33   2      0          2        0   150
#>   [3,]  35   5      1          2        0   172
#>   [4,]  35  10      1          9        0   106
#>   [5,]  34   5      0          5        0   120
#>   [6,]  35   2      0        180        0   121
#>   [7,]  37   9      0        180        1   151
#>   [8,]  38  13      1          0        1   161
#>   [9,]  38   2      0        115        0   150
#>  [10,]  36   1      0        180        1   155
#>  [11,]  35   0      0        180        1   119
#>  [12,]  38  12      1          8        1   120
#>  [13,]  36   5      1          0        1   115
#>  [14,]  38  12      1         11        1    92
#>  [15,]  40   2      1          1        1   148
#>  [16,]  38   5      1          3        0   125
#>  [17,]  42   2      0          2        0   140
#>  [18,]  40   6      0        180        1   138
#>  [19,]  40  11      1         10        1   120
#>  [20,]  41   2      1          1        0   166
#>  [21,]  40   1      1          0        1   145
#>  [22,]  43   4      1          0        1   130
#>  [23,]  40   3      1          1        0   170
#>  [24,]  42  12      1         10        1   170
#>  [25,]  43   2      1          1        1   116
#>  [26,]  44   5      1          1        0   170
#>  [27,]  45   3      0        180        1   154
#>  [28,]  41  10      1          8        0   150
#>  [29,]  44   3      0        180        0   141
#>  [30,]  45   9      1          7        0   110
#>  [31,]  45   6      0        180        1   170
#>  [32,]  41   5      1          4        1   141
#>  [33,]  44   2      1          1        1   150
#>  [34,]  45   2      0        180        1   140
#>  [35,]  46  15      0        180        0   120
#>  [36,]  46   2      1          1        0   126
#>  [37,]  48  15      0        180        1   160
#>  [38,]  45   3      0        150        0   130
#>  [39,]  44   3      1          0        1   180
#>  [40,]  46   7      1          2        0   166
#>  [41,]  43  29      0        180        1   180
#>  [42,]  45   4      1          0        0   124
#>  [43,]  43  10      0        180        0   185
#>  [44,]  47   6      1          0        1   116
#>  [45,]  46  13      1         10        0   100
#>  [46,]  44   0      1          0        1    96
#>  [47,]  47   4      1          3        1   160
#>  [48,]  43   3      1          0        1   124
#>  [49,]  45   8      1          0        1   117
#>  [50,]  49   5      0         73        1   136
#>  [51,]  46   2      1          1        1   122
#>  [52,]  44   4      1          0        1   114
#>  [53,]  47   2      0        180        0   108
#>  [54,]  44   9      1          8        1   135
#>  [55,]  45   5      0        180        1   190
#>  [56,]  45   9      1          0        1   145
#>  [57,]  47   3      1          1        1   120
#>  [58,]  48   3      0        180        0   154
#>  [59,]  48  12      1         11        0   200
#>  [60,]  47   9      1          6        0   170
#>  [61,]  46   3      1          0        1   119
#>  [62,]  47  10      0         10        1   140
#>  [63,]  50   1      1          0        1   129
#>  [64,]  48   2      1          0        0   184
#>  [65,]  47   7      0        180        0   145
#>  [66,]  50   6      1          2        1   140
#>  [67,]  49   7      1          7        1   110
#>  [68,]  46   3      1          1        1   140
#>  [69,]  51   1      0          1        1   145
#>  [70,]  47   2      0        180        0   150
#>  [71,]  46   6      1          0        1   156
#>  [72,]  52   2      0        180        1   170
#>  [73,]  50   7      1          0        1    92
#>  [74,]  50   4      0          4        1   100
#>  [75,]  51   3      1          2        0   113
#>  [76,]  50   1      1          0        0   150
#>  [77,]  50   9      0        180        0   130
#>  [78,]  49   7      1          4        1    90
#>  [79,]  47   6      0        180        1   162
#>  [80,]  51   8      0        180        1   140
#>  [81,]  52   2      0        180        0   155
#>  [82,]  46   3      0        180        1   120
#>  [83,]  50   4      1          1        0   150
#>  [84,]  48   7      1          0        1   110
#>  [85,]  53   8      0         36        1   160
#>  [86,]  47   2      1          1        0   110
#>  [87,]  52   4      1          4        0   152
#>  [88,]  49   9      1          3        0   102
#>  [89,]  54  17      1         12        1   102
#>  [90,]  53   7      1          0        0   199
#>  [91,]  54   6      1          3        0   129
#>  [92,]  51   3      1          1        0   140
#>  [93,]  50  10      1          6        0   122
#>  [94,]  50  14      1         13        0   170
#>  [95,]  53   8      1          7        0   160
#>  [96,]  48   3      1          2        0   150
#>  [97,]  49   5      1          2        1   150
#>  [98,]  52  14      1          7        1   200
#>  [99,]  48   6      0        180        0   160
#> [100,]  48  11      1         10        0   120
#> [101,]  53   4      1          0        1   156
#> [102,]  51  13      0         99        1   160
#> [103,]  54   9      1          0        1   138
#> [104,]  49  16      0         16        0   125
#> [105,]  55   6      1          2        1   114
#> [106,]  54   9      1          1        0   130
#> [107,]  51  13      1         11        0   145
#> [108,]  50   5      1          4        1   150
#> [109,]  55  28      1         13        1   160
#> [110,]  49   6      1          0        1   130
#> [111,]  49   1      0          1        1   110
#> [112,]  53   9      0          9        1    95
#> [113,]  53   8      1          0        1   130
#> [114,]  50   7      1          0        1   127
#> [115,]  56   4      1          1        1   130
#> [116,]  55   1      0        180        0   127
#> [117,]  55   2      0          2        0   145
#> [118,]  54   1      0        180        0   162
#> [119,]  54   7      1          0        1   100
#> [120,]  56   3      0        180        1   193
#> [121,]  55   5      1          4        1   120
#> [122,]  55   6      0        180        0   170
#> [123,]  52  16      0         16        0   152
#> [124,]  52  16      1         14        0   170
#> [125,]  53  15      0         15        1    90
#> [126,]  55   6      0        180        1   100
#> [127,]  54   3      0        180        0   128
#> [128,]  56   3      0          8        1   139
#> [129,]  55   1      0          2        0   130
#> [130,]  57   3      0          3        0   120
#> [131,]  54   2      1          1        1   176
#> [132,]  57   5      1          3        1   138
#> [133,]  56   4      1          0        1   140
#> [134,]  52   2      0        180        0   140
#> [135,]  55  11      1          7        0   104
#> [136,]  53   3      1          0        1   200
#> [137,]  58   8      0          8        1   130
#> [138,]  55   3      1          1        1   156
#> [139,]  57   0      0          0        1   150
#> [140,]  53  21      1         13        1   130
#> [141,]  59   3      1          1        0   172
#> [142,]  53   7      1          0        1   120
#> [143,]  57  11      1         10        1   129
#> [144,]  55   5      0          5        1   131
#> [145,]  54   7      1          0        1   141
#> [146,]  56   4      0          4        0   164
#> [147,]  58   1      1          1        1   200
#> [148,]  59   9      1          1        1   125
#> [149,]  60  11      1          9        0   106
#> [150,]  59   3      0        180        0   120
#> [151,]  58   4      1          0        1   160
#> [152,]  57   2      0          2        1   120
#> [153,]  60   5      1          1        0   138
#> [154,]  58  11      1          9        1   124
#> [155,]  55   5      1          0        1   160
#> [156,]  57  10      1          9        0   103
#> [157,]  59   6      1          0        1   140
#> [158,]  59   5      0        180        1   155
#> [159,]  59   4      1          0        1   152
#> [160,]  58  26      1          0        1   189
#> [161,]  61   9      0          9        1   160
#> [162,]  58   4      1          3        0   120
#> [163,]  60   0      1          0        1    80
#> [164,]  61   4      1          3        0   151
#> [165,]  61   9      1          8        0   150
#> [166,]  58   1      0          1        1   100
#> [167,]  61  20      1         13        0   130
#> [168,]  57  13      1         10        0   110
#> [169,]  57   2      1          0        1   116
#> [170,]  58  10      0         10        1   150
#> [171,]  57   4      1          3        0   138
#> [172,]  57  11      0        180        1   150
#> [173,]  61   3      0         17        0   143
#> [174,]  57   3      1          2        0   120
#> [175,]  58  19      1         13        1   140
#> [176,]  56  18      1         11        1   165
#> [177,]  55   4      1          3        1   160
#> [178,]  58  11      0        172        1   135
#> [179,]  60  12      1          0        1   114
#> [180,]  55   9      1          7        1   135
#> [181,]  61  13      1         12        1   130
#> [182,]  59  11      1          8        1   190
#> [183,]  57  15      1         13        1   110
#> [184,]  59   5      1          2        0   182
#> [185,]  58   5      1          1        1   135
#> [186,]  59  10      0        180        0   160
#> [187,]  61   8      0         77        0   120
#> [188,]  58   8      1          5        0   152
#> [189,]  62  10      1          0        1   153
#> [190,]  61  18      0        170        0   140
#> [191,]  61  28      1          7        0   133
#> [192,]  58   8      1          3        1   150
#> [193,]  61   7      0          7        1   150
#> [194,]  60   7      0          7        0   147
#> [195,]  61   6      0          6        0   134
#> [196,]  57  12      1          9        1   120
#> [197,]  62   4      1          3        0   173
#> [198,]  58   2      0         30        0   202
#> [199,]  59  16      1          9        1   133
#> [200,]  61  13      0         13        0   120
#> [201,]  57  18      1          9        1    93
#> [202,]  61   5      0          5        1   160
#> [203,]  58  11      1          9        0   179
#> [204,]  62  17      1         10        1   180
#> [205,]  62   1      1          0        1   172
#> [206,]  58   7      0        180        1   150
#> [207,]  63   1      0        180        1   130
#> [208,]  63   4      1          3        0   222
#> [209,]  62   3      0        180        1   105
#> [210,]  63   4      0        180        1   190
#> [211,]  63  15      1         10        1   126
#> [212,]  64   4      0        180        0   130
#> [213,]  60  18      1         13        0   132
#> [214,]  59   8      0        180        1   140
#> [215,]  62   7      0          7        0   150
#> [216,]  59   1      0         22        1   162
#> [217,]  58   2      0        180        0   127
#> [218,]  59   4      0        180        0   196
#> [219,]  60   7      0          7        0   140
#> [220,]  59   5      1          1        0   148
#> [221,]  65  13      0        180        1   100
#> [222,]  63   1      0          1        0   130
#> [223,]  62   6      0          6        0   120
#> [224,]  63  12      1         10        0   200
#> [225,]  60   8      0         17        1   130
#> [226,]  64  12      1         11        0   160
#> [227,]  66   1      1          0        1   120
#> [228,]  64   6      1          0        1   140
#> [229,]  63  10      1          0        1   148
#> [230,]  63  14      1          9        0   123
#> [231,]  65  36      1         11        0   140
#> [232,]  63   4      1          3        0   162
#> [233,]  66   3      1          1        0   127
#> [234,]  61  10      1          2        1   194
#> [235,]  64  32      1          9        1   160
#> [236,]  63  12      1          9        0   114
#> [237,]  63   7      0        180        0   120
#> [238,]  66   5      1          0        1   110
#> [239,]  65  10      1          8        1   120
#> [240,]  64   0      0          0        1    90
#> [241,]  60   6      0        180        0   130
#> [242,]  61  12      1         11        0   154
#> [243,]  65   3      0        180        1   190
#> [244,]  63  16      1          7        1   110
#> [245,]  64   7      0        180        1   120
#> [246,]  66   6      1          1        1   130
#> [247,]  63  12      0         12        1   150
#> [248,]  62   3      1          1        1   199
#> [249,]  65   6      0          9        0   112
#> [250,]  65   3      1          0        1    80
#> [251,]  63   5      1          4        0   170
#> [252,]  63   2      1          1        0   180
#> [253,]  62  13      1         11        0   180
#> [254,]  67  11      0         11        1   100
#> [255,]  64   2      0          2        0   201
#> [256,]  66  18      1          5        0   142
#> [257,]  66  16      1         11        1   169
#> [258,]  62   9      0        180        0   145
#> [259,]  63   3      1          2        0   120
#> [260,]  63   2      1          0        0   140
#> [261,]  65   8      1          0        1   140
#> [262,]  67   6      0        180        1   170
#> [263,]  65  15      1         11        1   160
#> [264,]  68   5      1          4        1   150
#> [265,]  64  13      1         12        1   150
#> [266,]  64   6      1          0        1   125
#> [267,]  66   7      1          0        1   115
#> [268,]  66  13      1          0        0   118
#> [269,]  64  14      1         13        1   150
#> [270,]  64   0      0          0        1   148
#> [271,]  67   4      1          3        0   130
#> [272,]  66   6      1          0        1   140
#> [273,]  68   1      0        180        1   166
#> [274,]  64  10      1          9        1   110
#> [275,]  67   8      1          1        1   130
#> [276,]  63  10      0         16        1   160
#> [277,]  66  14      0        180        0   130
#> [278,]  64   1      0          1        1   120
#> [279,]  68  18      0        180        1   260
#> [280,]  63   8      1          1        1   162
#> [281,]  65  18      1          3        0   120
#> [282,]  63  10      0         18        1   130
#> [283,]  67  11      0         11        0   150
#> [284,]  68  14      0         79        0   172
#> [285,]  65  15      1         12        1   150
#> [286,]  66  11      1          0        0   100
#> [287,]  65   4      1          2        1   145
#> [288,]  69  12      0         15        1   140
#> [289,]  63   2      0        180        0   150
#> [290,]  65  11      1          6        0   130
#> [291,]  69  21      1         10        0   180
#> [292,]  66   9      1          8        0   130
#> [293,]  63   8      0        180        1   120
#> [294,]  68  14      1         13        1   140
#> [295,]  65   8      1          0        1    90
#> [296,]  66   3      0          3        1   138
#> [297,]  69   1      1          0        0   170
#> [298,]  68  10      1         10        1   150
#> [299,]  67   2      0        180        0   184
#> [300,]  65  10      1          1        1   148
#> [301,]  67  12      1         12        0   160
#> [302,]  69   6      0         99        1   140
#> [303,]  64   4      0        179        0   160
#> [304,]  64  11      0         11        0   125
#> [305,]  64   0      1          0        1   118
#> [306,]  67   2      0         18        0   131
#> [307,]  66   7      1          5        1   131
#> [308,]  68   4      1          0        1   160
#> [309,]  69   4      1          3        1   150
#> [310,]  65  13      1         12        1   130
#> [311,]  66   6      0        180        0   140
#> [312,]  68  18      1          0        1   160
#> [313,]  65   6      0        101        1   115
#> [314,]  68   4      0          4        1   190
#> [315,]  71   3      0          5        0   112
#> [316,]  68   7      0        150        0   210
#> [317,]  71  20      1          0        1   160
#> [318,]  66   1      1          1        1   165
#> [319,]  70   4      1          0        1   180
#> [320,]  66   4      0        180        0   130
#> [321,]  67  10      1          9        0   200
#> [322,]  67   6      1          4        0   130
#> [323,]  68  18      1         14        1   170
#> [324,]  69   0      0          0        1   148
#> [325,]  68   7      1          0        1   150
#> [326,]  67  14      1         13        0   130
#> [327,]  65  14      1         13        1   150
#> [328,]  66   2      0          2        1   228
#> [329,]  71   6      0         45        1   158
#> [330,]  71   3      0        103        0   133
#> [331,]  70  22      1         13        0   103
#> [332,]  67   5      0          5        0   130
#> [333,]  68   6      0        180        0   145
#> [334,]  69   8      1          5        1   195
#> [335,]  69   6      1          4        1   174
#> [336,]  72   3      1          0        1   132
#> [337,]  66   2      1          1        0   123
#> [338,]  69  19      0        180        0   130
#> [339,]  68  18      0         18        1   100
#> [340,]  67  14      0        172        1   140
#> [341,]  69  11      1          0        1   120
#> [342,]  69   4      1          3        0   132
#> [343,]  68   2      0          7        1   130
#> [344,]  69   8      1          2        0   121
#> [345,]  72   5      1          4        0   170
#> [346,]  67  22      1          1        1   140
#> [347,]  68   3      0         19        0   135
#> [348,]  69   1      0          1        1   110
#> [349,]  67   4      0         60        1   136
#> [350,]  72  13      1         11        1   195
#> [351,]  66  24      1         13        0   130
#> [352,]  70  35      1          0        1   105
#> [353,]  72  30      1          0        1   145
#> [354,]  70   7      0          7        0   102
#> [355,]  73  20      1          0        1   170
#> [356,]  70  11      0        180        1   210
#> [357,]  72  19      1          8        0   120
#> [358,]  72  12      1         10        0   170
#> [359,]  67   5      1          0        1   147
#> [360,]  67   9      0        180        0   158
#> [361,]  70   5      0        180        0   150
#> [362,]  70   3      0        180        0   121
#> [363,]  69   3      0        180        0   220
#> [364,]  71   3      1          2        0   150
#> [365,]  68   4      1          3        0   210
#> [366,]  73   6      0        180        1   117
#> [367,]  69  16      1         10        1   140
#> [368,]  69   8      1          1        0   164
#> [369,]  72  16      1          1        1   130
#> [370,]  70   4      0        180        0   180
#> [371,]  69   1      1          0        0   155
#> [372,]  73   6      1          0        1   270
#> [373,]  72   8      1          1        1   150
#> [374,]  73   7      0          7        1   140
#> [375,]  68  15      1         13        1   130
#> [376,]  70   3      0          3        1   159
#> [377,]  70  13      1          9        0   100
#> [378,]  72   6      0        180        1   130
#> [379,]  73   0      0        180        1   161
#> [380,]  74   8      1          0        1    85
#> [381,]  73   4      0        180        1   154
#> [382,]  69   2      1          0        1   110
#> [383,]  71   3      1          1        0   150
#> [384,]  68   9      0        180        1   120
#> [385,]  71  20      1         10        0   140
#> [386,]  73   3      1          0        1   136
#> [387,]  70   5      1          0        1   190
#> [388,]  71   3      1          2        1   190
#> [389,]  69  12      1          1        1   149
#> [390,]  70  26      1         11        1   120
#> [391,]  74   4      0          4        0   120
#> [392,]  73   4      0         58        1   160
#> [393,]  72   5      1          3        1   160
#> [394,]  70   3      0        180        1   154
#> [395,]  73   6      0        180        0   110
#> [396,]  72   8      1          0        1   140
#> [397,]  71  13      1          8        0   121
#> [398,]  69   2      1          1        1    80
#> [399,]  70   4      1          0        1   140
#> [400,]  71  14      1         13        1   170
#> [401,]  74   7      1          0        1   117
#> [402,]  72  15      1         13        0   156
#> [403,]  71  10      1          9        1   120
#> [404,]  75   1      0          1        0   133
#> [405,]  75   2      1          1        0   145
#> [406,]  73  10      1          9        1   146
#> [407,]  73  10      1         10        1   120
#> [408,]  74  15      1          9        1   179
#> [409,]  71   2      0         10        1   112
#> [410,]  75   9      1          7        0   140
#> [411,]  75  13      1          1        1   130
#> [412,]  71  11      1          8        0   110
#> [413,]  71   4      0          4        0   134
#> [414,]  72  15      1         12        1   120
#> [415,]  70   7      1          4        0   184
#> [416,]  72   1      1          1        0   168
#> [417,]  72   7      0         57        1   145
#> [418,]  70   3      0          3        0   150
#> [419,]  73   5      1          3        1   112
#> [420,]  73  12      1         12        1   140
#> [421,]  72   2      0        180        0   120
#> [422,]  75   1      0        180        1   140
#> [423,]  71   3      1          0        0   144
#> [424,]  73   4      0        180        0   124
#> [425,]  74  34      1          8        1   233
#> [426,]  76   3      1          0        1   120
#> [427,]  71  32      1         12        1   107
#> [428,]  72   5      0        180        0   154
#> [429,]  76   5      0          5        1   130
#> [430,]  77  11      0         11        1   150
#> [431,]  77   4      0          4        0   185
#> [432,]  75   3      1          1        0   180
#> [433,]  72   7      1          2        0   142
#> [434,]  71  16      0        180        0   140
#> [435,]  73  10      1         10        0   124
#> [436,]  74   7      0        180        1   150
#> [437,]  74   3      0          3        1   128
#> [438,]  76   1      0        180        0   114
#> [439,]  73   6      0          6        1   114
#> [440,]  75  23      1         14        1   110
#> [441,]  74   2      0        180        0   190
#> [442,]  72   4      0         85        1   120
#> [443,]  72   4      1          3        0   160
#> [444,]  76  17      1          0        1   200
#> [445,]  73   4      1          3        1   125
#> [446,]  76  13      1         10        0   110
#> [447,]  75   7      0          7        0   190
#> [448,]  75  12      0         12        1   160
#> [449,]  76  13      1          8        1   148
#> [450,]  74   6      0        180        0   160
#> [451,]  76   4      0          4        1   155
#> [452,]  74   2      0        180        0   111
#> [453,]  73   1      0         52        1   105
#> [454,]  73   0      0        180        0   156
#> [455,]  72   5      0        180        0   120
#> [456,]  78  12      1         11        1   160
#> [457,]  76  44      1         10        0   105
#> [458,]  76   5      1          0        1   167
#> [459,]  74   8      1          8        1   170
#> [460,]  75   9      0        180        1   140
#> [461,]  73  33      1         12        1   175
#> [462,]  77   5      1          0        0   123
#> [463,]  77  12      0        180        0   130
#> [464,]  77   1      1          0        1    90
#> [465,]  76  12      1         11        1   120
#> [466,]  78   5      1          0        1   170
#> [467,]  73   7      1          0        0   174
#> [468,]  74   6      0         79        1   140
#> [469,]  74   9      1          8        0   200
#> [470,]  79  10      1          8        0   190
#> [471,]  74   2      1          0        1   130
#> [472,]  78  18      0         18        1   144
#> [473,]  77   3      0        180        0   110
#> [474,]  76  29      0         47        0    90
#> [475,]  73   8      1          1        1   162
#> [476,]  73  11      1          2        1   110
#> [477,]  74   2      0        180        0   100
#> [478,]  78   7      0          7        1   133
#> [479,]  74  15      0        180        1   172
#> [480,]  74   7      0          7        0   161
#> [481,]  76  13      1          1        1   170
#> [482,]  78  32      1          9        1   198
#> [483,]  79   6      0        180        0   170
#> [484,]  80  10      1          6        1   147
#> [485,]  78   0      0        180        1   212
#> [486,]  75   5      0        119        1   150
#> [487,]  75  12      1          1        1   120
#> [488,]  75  13      1          6        0   150
#> [489,]  74  10      1          8        0   135
#> [490,]  79   4      0         80        0   145
#> [491,]  78  12      1          9        0   150
#> [492,]  78   2      1          1        0   130
#> [493,]  75   4      1          0        0   212
#> [494,]  76  11      1          0        0   120
#> [495,]  75  11      1          4        0   162
#> [496,]  76   7      0         29        1   150
#> [497,]  77  24      0         24        1   160
#> [498,]  79   8      0         32        1   120
#> [499,]  78   6      1          0        1   240
#> [500,]  76   3      1          0        1   140
#> [501,]  78  11      1          1        1   140
#> [502,]  79  11      0        180        0   160
#> [503,]  78  14      1          0        1   140
#> [504,]  81   1      0          1        0   130
#> [505,]  79   4      0          4        1   125
#> [506,]  76  10      1          8        0   180
#> [507,]  76  12      1         10        1   127
#> [508,]  77   6      0          6        1   107
#> [509,]  80   3      1          0        1   120
#> [510,]  75   2      1          1        1   204
#> [511,]  78  11      0        180        1   135
#> [512,]  77  31      1          3        1   161
#> [513,]  76   1      0          1        1    90
#> [514,]  79   3      0          3        0   120
#> [515,]  77   7      0        180        1   170
#> [516,]  77   6      0          6        1   144
#> [517,]  81   1      0        180        0   120
#> [518,]  82   5      0          8        1   120
#> [519,]  80  40      1          0        1   138
#> [520,]  78   4      0         59        1   112
#> [521,]  80  17      1         12        0   100
#> [522,]  79  10      0         10        1   120
#> [523,]  81   4      1          2        1   126
#> [524,]  79  28      0        164        0   100
#> [525,]  80   6      0        173        1   160
#> [526,]  81   3      0        180        0   184
#> [527,]  81   2      0        175        0   172
#> [528,]  78   7      0          7        1   147
#> [529,]  78  15      0         15        0   165
#> [530,]  78   4      0        180        0   175
#> [531,]  78  26      1          5        0   194
#> [532,]  76   1      0        166        0   131
#> [533,]  81   4      1          1        1   104
#> [534,]  80   1      0          1        0   100
#> [535,]  78   3      1          1        1   152
#> [536,]  77  10      1          8        1   130
#> [537,]  82   3      1          1        1   144
#> [538,]  80   2      1          1        0   168
#> [539,]  79   6      0          6        0   152
#> [540,]  80   6      1          0        1   119
#> [541,]  78   2      0        180        0   148
#> [542,]  79  10      0        180        1   150
#> [543,]  77   4      0        180        1    98
#> [544,]  81   1      0        108        0   129
#> [545,]  78  12      0        180        0   134
#> [546,]  79   1      0        125        0   193
#> [547,]  84  22      1         10        0   180
#> [548,]  79   4      0          4        1   121
#> [549,]  80   6      0          6        1   110
#> [550,]  83   9      1          5        1   170
#> [551,]  82   5      0        180        0   110
#> [552,]  83   5      0        180        0   148
#> [553,]  83   4      0        103        0    97
#> [554,]  81   5      0        177        0    41
#> [555,]  80  11      1          8        0   170
#> [556,]  78  23      1         10        1   145
#> [557,]  79   4      0          4        1   183
#> [558,]  78   9      1          4        1   120
#> [559,]  79   1      0        180        1   170
#> [560,]  81  15      0        180        1   140
#> [561,]  80   7      1          0        1   146
#> [562,]  81  20      1          9        0   170
#> [563,]  83   8      0          8        0   115
#> [564,]  81  16      0         16        1   110
#> [565,]  80  11      1          8        0   110
#> [566,]  80   8      1          7        0   160
#> [567,]  79   0      1          0        1    96
#> [568,]  85   4      0        180        0    90
#> [569,]  83   2      0          2        1   155
#> [570,]  82   6      0        128        1   100
#> [571,]  80   3      1          1        1   230
#> [572,]  82  23      1          0        0   110
#> [573,]  81   1      0          1        1   150
#> [574,]  83   3      0        180        0   174
#> [575,]  81   4      0         90        1   138
#> [576,]  80  13      1          8        1   140
#> [577,]  80   2      1          0        1   130
#> [578,]  79   4      0          4        1    60
#> [579,]  80   6      0         71        1   189
#> [580,]  82  19      0         19        0   120
#> [581,]  83   9      0        180        0   198
#> [582,]  79  14      1          0        0   110
#> [583,]  83   3      0        114        0    98
#> [584,]  81  14      1         12        1   128
#> [585,]  83   2      0        154        0   130
#> [586,]  82   0      0          2        1   100
#> [587,]  85   9      1          6        1   160
#> [588,]  83   1      0        180        0   160
#> [589,]  81   1      0          1        1   145
#> [590,]  81  12      0         12        1   163
#> [591,]  82  16      1          8        0   103
#> [592,]  82   5      1          0        1   146
#> [593,]  81  19      1         14        0   120
#> [594,]  82   3      1          2        0   130
#> [595,]  83   7      0        126        0   135
#> [596,]  86   8      0          8        1   132
#> [597,]  81  16      1          9        0   180
#> [598,]  84   6      0        165        0   145
#> [599,]  82   9      0        180        1   134
#> [600,]  84   3      0        180        1   120
#> [601,]  81  13      0        180        0   152
#> [602,]  81   2      1          0        1   118
#> [603,]  81   4      0        180        0   160
#> [604,]  83   9      0        180        1   149
#> [605,]  82   1      0        180        1   193
#> [606,]  86  12      1          0        1   132
#> [607,]  86   6      1          0        1   140
#> [608,]  83  19      0         43        0   150
#> [609,]  83  10      1          0        1   190
#> [610,]  86   2      0        180        1   169
#> [611,]  84   3      0          3        1   121
#> [612,]  86   4      0         38        1   122
#> [613,]  82   4      0          4        0   130
#> [614,]  86   6      0          6        1   117
#> [615,]  84  13      0         62        1   100
#> [616,]  86   6      1          1        0   112
#> [617,]  88   4      0          4        0   100
#> [618,]  83  20      1          3        1   150
#> [619,]  85  22      0         22        1   184
#> [620,]  86   9      1          7        1   142
#> [621,]  87   2      0        180        1   130
#> [622,]  86   6      0         46        0   173
#> [623,]  88   2      0        180        1    68
#> [624,]  83   3      0          3        1   130
#> [625,]  86  15      1          8        1   109
#> [626,]  88   4      0          4        0    86
#> [627,]  89   4      0          4        1   153
#> [628,]  89   5      0        119        1   140
#> [629,]  87   6      0        180        1   110
#> [630,]  87   1      0          1        0   170
#> [631,]  84   8      0        180        1   119
#> [632,]  85   8      0          8        1   136
#> [633,]  87  29      0         29        1    97
#> [634,]  87  15      1          9        1   138
#> [635,]  89  10      0         46        1   170
#> [636,]  90  14      0         14        1   100
#> [637,]  88   1      0          1        0   135
#> [638,]  86   4      0        180        1   145
#> [639,]  91   8      0          8        0   100
#> [640,]  87   2      0        180        0   160
#> [641,]  91  10      0        145        0   135
#> [642,]  88   7      0         24        0   119
#> [643,]  88   8      0         50        1   154
#> [644,]  86  10      0        180        1   137
#> [645,]  86   9      1          7        0   130
#> [646,]  90   4      1          0        0   121
#> [647,]  91   1      0          1        1    74
#> [648,]  87  43      0        178        1   130
#> [649,]  91   3      0         33        1   137
#> [650,]  87   7      0         74        1   105
#> [651,]  89  12      1          0        1   130
#> [652,]  89   2      0        168        0   118
#> [653,]  92   7      0          7        1   110
#> [654,]  91   0      0          0        0     0
#> [655,]  89  14      0        180        1    84
#> [656,]  90  18      0        180        0   188
#> [657,]  91   4      1          0        1   120
#> [658,]  90  19      1         11        1   129
#> [659,]  91   2      0          2        1   116
#> [660,]  94   8      0          8        1   142
#> [661,]  92   4      0         76        1   149
#> [662,]  91   1      0        180        0   158
#> [663,]  90  16      0         16        1   106
#> [664,]  90   3      0         67        0   162
#> [665,]  94   3      0         26        1   144
#> [666,]  91  12      0         53        1   212
#> [667,]  91   7      0          7        0   135
#> [668,]  92   5      0         69        0   139
#> [669,]  92   2      0          2        0   112
#> [670,]  96  15      1          0        1   140
#> 
#> $y
#>   [1] 180.0+   2.0+   5.0+ 180.0+   5.0+ 180.0+ 180.0+ 180.0+ 115.0  180.0+
#>  [11] 180.0+  12.0    5.0+ 180.0+   2.0+   5.0+   2.0+ 180.0+ 180.0+ 180.0+
#>  [21] 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 155.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [31] 180.0+   5.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 150.0  180.0+ 180.0+
#>  [41] 180.0+ 180.0+ 180.0+   6.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  73.0 
#>  [51] 161.0+ 180.0+ 180.0+ 180.0+ 180.0+ 177.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [61] 180.0+  10.0+ 172.0+ 180.0+ 180.0+ 180.0+   7.0  180.0+   1.0  180.0+
#>  [71] 180.0+ 180.0+ 180.0+   4.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [81] 180.0+ 180.0+ 180.0+   7.0   36.0  180.0+   4.0+ 180.0+ 180.0+ 180.0+
#>  [91] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  85.0  180.0+ 180.0+
#> [101] 166.0+  99.0  180.0+  16.0+   6.0+ 180.0+  13.0+ 171.0+  28.0    6.0+
#> [111]   1.0    9.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0  180.0+   7.0+ 180.0+
#> [121] 180.0+ 180.0+  16.0+  16.0   15.0+ 180.0+ 180.0+   8.0    2.0    3.0+
#> [131] 180.0+ 140.0  165.0  180.0+ 180.0+ 180.0+   8.0+ 180.0+   0.5  180.0+
#> [141] 180.0+ 180.0+ 180.0+   5.0+ 180.0+   4.0+   1.0  180.0+ 180.0+ 180.0+
#> [151] 180.0+   2.0  180.0+ 180.0+ 180.0+ 180.0+  64.0  180.0+ 180.0+ 180.0+
#> [161]   9.0+ 180.0+   0.5  180.0+ 180.0+   1.0  180.0+ 180.0+ 180.0+  10.0+
#> [171] 180.0+ 180.0+  17.0    3.0+  19.0  180.0+ 180.0+ 172.0+ 172.0+  24.0 
#> [181] 180.0+ 180.0+  15.0  180.0+ 180.0+ 180.0+  77.0    8.0+ 180.0+ 170.0 
#> [191]  94.0  180.0+   7.0    7.0+   6.0  180.0+ 180.0+  30.0  180.0+  13.0+
#> [201]  18.0    5.0+ 180.0+ 180.0+   1.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [211] 180.0+ 180.0+ 180.0+ 180.0+   7.0+  22.0  180.0+ 180.0+   7.0+ 180.0+
#> [221] 180.0+   1.0    6.0+ 180.0+  17.0   12.0  180.0+ 180.0+ 180.0+  14.0+
#> [231]  36.0  180.0+   3.0+  88.0  180.0+  12.0  180.0+ 180.0+ 180.0+   0.5 
#> [241] 180.0+  12.0+ 180.0+ 180.0+ 180.0+ 180.0+  12.0  180.0+   9.0    3.0 
#> [251] 180.0+ 180.0+ 180.0+  11.0+   2.0+  18.0+ 180.0+ 180.0+   3.0+   2.0+
#> [261]  15.0  180.0+ 180.0+   5.0+  13.0  180.0+ 179.0+ 166.0+  14.0+   0.5+
#> [271] 180.0+ 180.0+ 180.0+ 180.0+   8.0   16.0  180.0+   1.0  180.0+ 180.0+
#> [281] 123.0+  18.0   11.0+  79.0   15.0+ 180.0+   4.0+  15.0  180.0+ 180.0+
#> [291] 174.0+ 180.0+ 180.0+ 180.0+   8.0+   3.0  175.0   10.0  180.0+ 180.0+
#> [301]  12.0   99.0  179.0+  11.0+   0.5   18.0    7.0+ 180.0+ 152.0+ 180.0+
#> [311] 180.0+  18.0+ 101.0    4.0    5.0  150.0  180.0+   1.0  180.0+ 180.0+
#> [321] 174.0+   6.0  180.0+   0.5  180.0+ 180.0+  14.0+   2.0   45.0  103.0 
#> [331] 180.0+   5.0+ 180.0+ 180.0+  97.0  180.0+   2.0+ 180.0+  18.0  172.0+
#> [341] 180.0+ 180.0+   7.0    8.0+ 180.0+  51.0   19.0    1.0   60.0  132.0 
#> [351] 180.0+ 180.0+ 162.0    7.0+ 124.0  180.0+ 180.0+  12.0  180.0+ 180.0+
#> [361] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  16.0+ 180.0+  16.0+ 180.0+
#> [371] 180.0+   6.0  180.0+   7.0+  15.0    3.0+  13.0+ 180.0+ 180.0+ 180.0+
#> [381] 180.0+   2.0    3.0+ 180.0+  20.0  180.0+ 180.0+   3.0   12.0  180.0+
#> [391]   4.0+  58.0  180.0+ 180.0+ 180.0+ 180.0+ 175.0    2.0  180.0+  14.0+
#> [401] 180.0+ 180.0+ 179.0+   1.0  180.0+ 180.0+  15.0  180.0+  10.0  180.0+
#> [411]  13.0  180.0+   4.0+ 180.0+ 104.0+   1.0   57.0    3.0+   5.0   12.0 
#> [421] 180.0+ 180.0+ 180.0+ 180.0+  34.0  180.0+ 177.0+ 180.0+   5.0   11.0+
#> [431]   4.0+ 180.0+   7.0  180.0+  10.0  180.0+   3.0  180.0+   6.0  180.0+
#> [441] 180.0+  85.0  180.0+  17.0+ 180.0+ 174.0+   7.0   12.0  180.0+ 180.0+
#> [451]   4.0  180.0+  52.0  180.0+ 180.0+  12.0  180.0+ 180.0+   8.0  180.0+
#> [461]  33.0    5.0  180.0+   1.0   12.0  180.0+   7.0+  79.0  168.0+ 180.0+
#> [471] 176.0+  18.0  180.0+  47.0  180.0+  11.0  180.0+   7.0  180.0+   7.0 
#> [481] 180.0+  32.0  180.0+  10.0  180.0+ 119.0   12.0  180.0+ 180.0+  80.0 
#> [491] 180.0+ 180.0+   4.0+  11.0  152.0+  29.0   24.0   32.0  180.0+   3.0+
#> [501] 180.0+ 180.0+ 180.0+   1.0    4.0   10.0+ 180.0+   6.0    3.0+   2.0+
#> [511] 180.0+ 171.0    1.0    3.0  180.0+   6.0  180.0+   8.0   40.0   59.0 
#> [521]  17.0   10.0+  93.0  164.0  173.0  180.0+ 175.0+   7.0+  15.0+ 180.0+
#> [531] 171.0+ 166.0+  71.0    1.0    3.0+  10.0  180.0+  10.0    6.0+   6.0 
#> [541] 180.0+ 180.0+ 180.0+ 108.0  180.0+ 125.0  180.0+   4.0    6.0    9.0+
#> [551] 180.0+ 180.0+ 103.0  177.0+ 169.0   70.0    4.0  180.0+ 180.0+ 180.0+
#> [561]   7.0+  20.0    8.0+  16.0  180.0+ 180.0+   0.5  180.0+   2.0  128.0 
#> [571]   3.0+  62.0    1.0  180.0+  90.0  180.0+ 180.0+   4.0   71.0   19.0 
#> [581] 180.0+ 180.0+ 114.0  180.0+ 154.0    2.0  180.0+ 180.0+   1.0   12.0 
#> [591]  16.0+   5.0+ 180.0+   3.0  126.0    8.0  180.0+ 165.0  180.0+ 180.0+
#> [601] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   6.0   43.0  180.0+ 180.0+
#> [611]   3.0   38.0    4.0    6.0+  62.0    6.0+   4.0+  20.0   22.0   11.0 
#> [621] 180.0+  46.0  180.0+   3.0+ 180.0+   4.0    4.0  119.0  180.0+   1.0+
#> [631] 180.0+   8.0   29.0  180.0+  46.0   14.0    1.0+ 180.0+   8.0  180.0+
#> [641] 145.0   24.0   50.0  180.0+ 180.0+   4.0    1.0  178.0+  33.0   74.0 
#> [651] 180.0+ 168.0    7.0    0.5  180.0+ 180.0+   4.0  180.0+   2.0    8.0+
#> [661]  76.0  180.0+  16.0   67.0   26.0   53.0    7.0+  69.0    2.0   15.0+
#> 
#> $weights
#> NULL
#> 
#> $offset
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Warning: Multiple lambdas have been fit. Lambda will be set to 0.01 (see parameter 's').

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8591781