GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.coxnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.coxnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Caution: This learner is different to learners calling glmnet::cv.glmnet() in that it does not use the internal optimization of parameter lambda. Instead, lambda needs to be tuned by the user (e.g., via mlr3tuning). When lambda is tuned, the glmnet will be trained for each tuning iteration. While fitting the whole path of lambdas would be more efficient, as is done by default in glmnet::glmnet(), tuning/selecting the parameter at prediction time (using parameter s) is currently not supported in mlr3 (at least not in efficient manner). Tuning the s parameter is, therefore, currently discouraged.

When the data are i.i.d. and efficiency is key, we recommend using the respective auto-tuning counterpart in mlr_learners_surv.cv_glmnet(). However, in some situations this is not applicable, usually when data are imbalanced or not i.i.d. (longitudinal, time-series) and tuning requires custom resampling strategies (blocked design, stratification).

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.coxnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exact	logical	FALSE	TRUE, FALSE	-

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	TRUE	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	0.01		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.logistic	character	Newton	Newton, modified.Newton	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvGlmnet

Methods

Public methods

LearnerSurvGlmnet$new()

LearnerSurvGlmnet$selected_features()

LearnerSurvGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.glmnet")
print(learner)
#> <LearnerSurvGlmnet:surv.glmnet>: Regularized Generalized Linear Model
#> * Model: -
#> * Parameters: use_pred_offset=TRUE
#> * Packages: mlr3, mlr3proba, mlr3extralearners, glmnet
#> * Predict Types:  [crank], distr, lp
#> * Feature Types: logical, integer, numeric
#> * Properties: offset, selected_features, weights

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#>    Df  %Dev   Lambda
#> 1   0  0.00 0.202800
#> 2   1  0.64 0.184800
#> 3   1  1.18 0.168400
#> 4   2  1.98 0.153400
#> 5   2  2.80 0.139800
#> 6   3  5.05 0.127400
#> 7   3  7.51 0.116100
#> 8   3  9.35 0.105800
#> 9   3 10.79 0.096360
#> 10  3 11.93 0.087800
#> 11  3 12.86 0.080000
#> 12  3 13.62 0.072900
#> 13  4 14.26 0.066420
#> 14  4 14.82 0.060520
#> 15  4 15.29 0.055140
#> 16  4 15.69 0.050240
#> 17  5 16.04 0.045780
#> 18  5 16.35 0.041710
#> 19  5 16.61 0.038010
#> 20  6 16.83 0.034630
#> 21  6 17.03 0.031560
#> 22  6 17.19 0.028750
#> 23  6 17.33 0.026200
#> 24  6 17.44 0.023870
#> 25  6 17.54 0.021750
#> 26  6 17.62 0.019820
#> 27  6 17.69 0.018060
#> 28  6 17.75 0.016450
#> 29  6 17.80 0.014990
#> 30  6 17.84 0.013660
#> 31  6 17.87 0.012450
#> 32  6 17.90 0.011340
#> 33  6 17.93 0.010330
#> 34  6 17.95 0.009415
#> 35  6 17.97 0.008579
#> 36  6 17.98 0.007816
#> 37  6 17.99 0.007122
#> 38  6 18.00 0.006489
#> 39  6 18.01 0.005913
#> 40  6 18.02 0.005388
#> 41  6 18.02 0.004909
#> 42  6 18.03 0.004473
#> 43  6 18.03 0.004075
#> 44  6 18.03 0.003713
#> 45  6 18.04 0.003384
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  28   9      0        180        1   107
#>   [2,]  32   5      1          0        1   121
#>   [3,]  35   5      1          2        0   172
#>   [4,]  35  10      1          9        0   106
#>   [5,]  35   2      1          1        1   112
#>   [6,]  37   9      0        180        1   151
#>   [7,]  38  13      1          0        1   161
#>   [8,]  38   2      0        115        0   150
#>   [9,]  38  12      1          8        1   120
#>  [10,]  33   6      1          1        1   115
#>  [11,]  38  16      1         10        0   160
#>  [12,]  38  12      1         11        1    92
#>  [13,]  40  12      1          9        0   153
#>  [14,]  42   3      1          1        1   130
#>  [15,]  37   1      1          0        1   146
#>  [16,]  40   2      1          1        1   148
#>  [17,]  42   2      0        180        1   100
#>  [18,]  42   2      0          2        0   140
#>  [19,]  40  11      1         10        1   120
#>  [20,]  41   2      1          1        0   166
#>  [21,]  40   1      1          0        1   145
#>  [22,]  43   4      1          0        1   130
#>  [23,]  42   4      0        180        0   162
#>  [24,]  40   3      1          1        0   170
#>  [25,]  42  12      1         10        1   170
#>  [26,]  43   2      1          1        1   116
#>  [27,]  44   5      1          1        0   170
#>  [28,]  45   3      0        180        1   154
#>  [29,]  41  10      1          8        0   150
#>  [30,]  44   3      0        180        0   141
#>  [31,]  41  13      1          1        0   140
#>  [32,]  45   9      1          7        0   110
#>  [33,]  45   6      0        180        1   170
#>  [34,]  44   2      1          1        1   150
#>  [35,]  43   2      0        180        1   140
#>  [36,]  45   2      0        180        1   140
#>  [37,]  47   4      1          3        0   118
#>  [38,]  48  15      0        180        1   160
#>  [39,]  45   3      0        150        0   130
#>  [40,]  46   7      1          2        0   166
#>  [41,]  45   4      1          0        0   124
#>  [42,]  43  10      0        180        0   185
#>  [43,]  47   6      1          0        1   116
#>  [44,]  46  13      1         10        0   100
#>  [45,]  44   0      1          0        1    96
#>  [46,]  47   4      1          3        1   160
#>  [47,]  49   5      0         73        1   136
#>  [48,]  45   5      0          5        0   141
#>  [49,]  46   2      1          1        1   122
#>  [50,]  46   6      1          0        1   100
#>  [51,]  46   5      1          3        0   130
#>  [52,]  46   4      0        180        1   121
#>  [53,]  44   2      0        180        0   142
#>  [54,]  46  15      0        180        1   120
#>  [55,]  45   9      1          0        1   145
#>  [56,]  47   3      1          1        1   120
#>  [57,]  48   3      0        180        0   154
#>  [58,]  48  12      1         11        0   200
#>  [59,]  46   3      1          0        1   119
#>  [60,]  49   4      0        180        0   117
#>  [61,]  50   1      1          0        1   129
#>  [62,]  47   7      0        180        0   145
#>  [63,]  50   4      1          1        0   125
#>  [64,]  50   6      1          2        1   140
#>  [65,]  46   9      1          9        1   122
#>  [66,]  49   2      0          2        0   105
#>  [67,]  51   1      0          1        1   145
#>  [68,]  49  15      1         11        1   160
#>  [69,]  47   2      0        180        0   150
#>  [70,]  49  23      0        179        1   112
#>  [71,]  50   7      1          0        1    92
#>  [72,]  50   4      0          4        1   100
#>  [73,]  51   3      1          2        0   113
#>  [74,]  50   1      1          0        0   150
#>  [75,]  50   9      0        180        0   130
#>  [76,]  49   7      1          4        1    90
#>  [77,]  47   8      0        180        0   160
#>  [78,]  51   8      0        180        1   140
#>  [79,]  52   2      0        180        0   155
#>  [80,]  46   1      1          1        0   145
#>  [81,]  50   4      1          1        0   150
#>  [82,]  53   8      0         36        1   160
#>  [83,]  48  17      1         10        0   111
#>  [84,]  52   4      1          4        0   152
#>  [85,]  49   9      1          3        0   102
#>  [86,]  49  15      0        180        1   160
#>  [87,]  54  17      1         12        1   102
#>  [88,]  53   5      0         77        0   159
#>  [89,]  53   7      1          0        0   199
#>  [90,]  51   3      1          1        0   140
#>  [91,]  50   2      0          5        1   106
#>  [92,]  50  10      1          6        0   122
#>  [93,]  50  14      1         13        0   170
#>  [94,]  53   8      1          7        0   160
#>  [95,]  48   3      1          2        0   150
#>  [96,]  51  25      1          1        0   202
#>  [97,]  49   5      1          2        1   150
#>  [98,]  53   4      0          4        0   140
#>  [99,]  52  14      1          7        1   200
#> [100,]  48   6      0        180        0   160
#> [101,]  48  11      1         10        0   120
#> [102,]  53   4      1          0        1   156
#> [103,]  51  13      0         99        1   160
#> [104,]  54   9      1          0        1   138
#> [105,]  49  16      0         16        0   125
#> [106,]  55   3      1          1        0   150
#> [107,]  54   9      1          1        0   130
#> [108,]  55   4      1          2        0   150
#> [109,]  51  13      1         11        0   145
#> [110,]  50   5      1          4        1   150
#> [111,]  54   4      1          0        1   121
#> [112,]  50   3      0        174        1   153
#> [113,]  49   6      1          0        1   130
#> [114,]  49   1      0          1        1   110
#> [115,]  53   9      0          9        1    95
#> [116,]  50   7      1          0        1   127
#> [117,]  56   4      1          1        1   130
#> [118,]  54   7      1          0        1   100
#> [119,]  56   2      0        180        0   132
#> [120,]  55   5      1          4        1   120
#> [121,]  52   8      0        180        0   119
#> [122,]  53  18      1          9        1   150
#> [123,]  54   3      0        180        1   180
#> [124,]  55   6      0        180        0   170
#> [125,]  52  16      1         14        0   170
#> [126,]  53  15      0         15        1    90
#> [127,]  53   4      0        180        1   150
#> [128,]  55   6      0        180        1   100
#> [129,]  54  12      1          0        1   190
#> [130,]  56   3      0          8        1   139
#> [131,]  54   7      1          2        0   129
#> [132,]  54   2      1          1        0   135
#> [133,]  52   9      1          3        0   170
#> [134,]  54   2      1          1        1   176
#> [135,]  57   1      0          1        1   100
#> [136,]  56   4      1          0        1   140
#> [137,]  52   2      0        180        0   140
#> [138,]  52  15      1         14        0   130
#> [139,]  56  14      1         11        0   130
#> [140,]  53   3      1          0        1   200
#> [141,]  57  10      0        180        1   170
#> [142,]  54   5      0        180        1   108
#> [143,]  55   3      1          1        1   156
#> [144,]  57   0      0          0        1   150
#> [145,]  53  21      1         13        1   130
#> [146,]  59   3      1          1        0   172
#> [147,]  58   6      1          0        1    90
#> [148,]  56   5      0          5        1   150
#> [149,]  54  23      1          8        0   120
#> [150,]  57   4      1          2        1   185
#> [151,]  53   4      0        147        1   145
#> [152,]  57  11      1         10        1   129
#> [153,]  55   5      0          5        1   131
#> [154,]  54   7      1          0        1   141
#> [155,]  59  15      1         10        0   140
#> [156,]  58   1      1          1        1   200
#> [157,]  55   5      1          0        0   140
#> [158,]  56   7      1          5        1   120
#> [159,]  55   2      0          2        0   106
#> [160,]  59   9      1          1        1   125
#> [161,]  58   4      1          0        1   160
#> [162,]  57   2      0          2        1   120
#> [163,]  60   5      1          1        0   138
#> [164,]  59   6      1          0        1   140
#> [165,]  59   5      0        180        1   155
#> [166,]  58  26      1          0        1   189
#> [167,]  58   4      1          3        0   120
#> [168,]  60   0      1          0        1    80
#> [169,]  58   8      0        161        1   140
#> [170,]  58  14      1          6        0   190
#> [171,]  61   4      1          3        0   151
#> [172,]  61   9      1          8        0   150
#> [173,]  58   1      0          1        1   100
#> [174,]  57  13      1         10        0   110
#> [175,]  57   2      1          0        1   116
#> [176,]  58  10      0         10        1   150
#> [177,]  57   4      1          3        0   138
#> [178,]  57  11      0        180        1   150
#> [179,]  56  14      0         45        0   130
#> [180,]  57   3      1          2        0   120
#> [181,]  58  19      1         13        1   140
#> [182,]  56  18      1         11        1   165
#> [183,]  59   9      1          0        1    80
#> [184,]  55   4      1          3        1   160
#> [185,]  58  11      0        172        1   135
#> [186,]  60  12      1          0        1   114
#> [187,]  56   8      1          8        0   120
#> [188,]  61  13      1         12        1   130
#> [189,]  59  11      1          8        1   190
#> [190,]  57   1      0          1        0   126
#> [191,]  58   5      1          1        1   135
#> [192,]  61   8      0         77        0   120
#> [193,]  61  13      0         13        0   210
#> [194,]  57   3      1          0        0   100
#> [195,]  61  18      0        170        0   140
#> [196,]  61  28      1          7        0   133
#> [197,]  58   8      1          3        1   150
#> [198,]  61   7      0          7        1   150
#> [199,]  60   7      0          7        0   147
#> [200,]  59  13      1          2        0   198
#> [201,]  62   4      1          0        0   160
#> [202,]  60  17      1          8        1   140
#> [203,]  58   3      1          0        1   146
#> [204,]  61  13      0         13        0   120
#> [205,]  61   5      0          5        1   110
#> [206,]  58  11      1          9        0   179
#> [207,]  62  17      1         10        1   180
#> [208,]  62   1      1          0        1   172
#> [209,]  58   7      0        180        1   150
#> [210,]  63   3      1          1        0   180
#> [211,]  63   4      1          3        0   222
#> [212,]  62   3      0        180        1   105
#> [213,]  63   4      0        180        1   190
#> [214,]  63  15      1         10        1   126
#> [215,]  64   4      0        180        0   130
#> [216,]  63   4      1          1        0   155
#> [217,]  60  18      1         13        0   132
#> [218,]  59   8      0        180        1   140
#> [219,]  61   9      1          9        1   150
#> [220,]  58   9      1          9        0   110
#> [221,]  58   2      0        180        0   127
#> [222,]  60   7      1          5        1   141
#> [223,]  60   7      0          7        0   140
#> [224,]  59   5      1          1        0   148
#> [225,]  60   7      1          1        1    90
#> [226,]  65  13      0        180        1   100
#> [227,]  63   1      0          1        0   162
#> [228,]  63   1      0          1        0   130
#> [229,]  61  15      1         13        0   170
#> [230,]  60   3      0          3        0   168
#> [231,]  62   6      0          6        0   120
#> [232,]  63  12      1         10        0   200
#> [233,]  59  10      0        180        1   130
#> [234,]  61   6      1          1        1   117
#> [235,]  66   1      1          0        1   120
#> [236,]  64   6      1          0        1   140
#> [237,]  63  14      1          9        0   123
#> [238,]  65  36      1         11        0   140
#> [239,]  63   4      1          3        0   162
#> [240,]  66   3      1          1        0   127
#> [241,]  61  10      1          2        1   194
#> [242,]  64  32      1          9        1   160
#> [243,]  63   7      0        180        0   120
#> [244,]  66   5      1          0        1   110
#> [245,]  64   0      0          0        1    90
#> [246,]  64   9      0        180        0   150
#> [247,]  61   4      0        180        1   113
#> [248,]  65   3      0        180        1   190
#> [249,]  63  16      1          7        1   110
#> [250,]  66   6      1          1        1   130
#> [251,]  63  12      0         12        1   150
#> [252,]  62   3      1          1        1   199
#> [253,]  65   6      0          9        0   112
#> [254,]  65   3      1          0        1    80
#> [255,]  63   2      1          1        0   180
#> [256,]  62  13      1         11        0   180
#> [257,]  67  11      0         11        1   100
#> [258,]  64   2      0          2        0   201
#> [259,]  66  16      1         11        1   169
#> [260,]  62   9      0        180        0   145
#> [261,]  61  15      1         10        0   130
#> [262,]  63   9      1          8        1   160
#> [263,]  63   3      1          2        0   120
#> [264,]  63   2      1          0        0   140
#> [265,]  64  19      1          8        1   160
#> [266,]  65   8      1          0        1   140
#> [267,]  65  15      1         11        1   160
#> [268,]  64  13      1         12        1   150
#> [269,]  66   7      1          0        1   115
#> [270,]  66  13      1          0        0   118
#> [271,]  65   3      0          3        0   105
#> [272,]  64   0      0          0        1   148
#> [273,]  66   3      1          0        1   135
#> [274,]  66   6      1          0        1   140
#> [275,]  68   1      0        180        1   166
#> [276,]  64  10      1          9        1   110
#> [277,]  67   8      1          1        1   130
#> [278,]  63  10      0         16        1   160
#> [279,]  66  14      0        180        0   130
#> [280,]  65  17      1         14        1   100
#> [281,]  63   8      1          1        1   162
#> [282,]  65  18      1          3        0   120
#> [283,]  63   1      1          0        1   155
#> [284,]  68  11      0        180        0   160
#> [285,]  66  12      1         10        1   150
#> [286,]  65  15      1         12        1   150
#> [287,]  66  15      1         13        1   160
#> [288,]  63   2      0        180        0   150
#> [289,]  65  11      1          6        0   130
#> [290,]  69  21      1         10        0   180
#> [291,]  69   6      0        180        1   100
#> [292,]  63   8      0        180        1   120
#> [293,]  65   8      1          0        1    90
#> [294,]  66   3      0          3        1   138
#> [295,]  69   1      1          0        0   170
#> [296,]  67   1      0        180        1   160
#> [297,]  68  10      1         10        1   150
#> [298,]  67   7      1          4        1   130
#> [299,]  67   2      0        180        0   184
#> [300,]  65   6      0          6        0    80
#> [301,]  67  12      1         12        0   160
#> [302,]  69   6      0         99        1   140
#> [303,]  65   4      1          1        0   130
#> [304,]  64   4      0        179        0   160
#> [305,]  66   4      0        180        1   130
#> [306,]  64   4      0        180        1   140
#> [307,]  67   2      0         18        0   131
#> [308,]  66   7      1          5        1   131
#> [309,]  66   4      0        180        0   177
#> [310,]  69   8      0         93        0   140
#> [311,]  64  21      0         21        1   155
#> [312,]  66   6      0        180        0   140
#> [313,]  68  18      1          0        1   160
#> [314,]  65   6      0        101        1   115
#> [315,]  68   4      0          4        1   190
#> [316,]  71   3      0          5        0   112
#> [317,]  70   7      1          0        1   190
#> [318,]  67   2      0        180        0   128
#> [319,]  66   1      1          1        1   165
#> [320,]  70   4      1          0        1   180
#> [321,]  69   8      0        180        1   153
#> [322,]  70  14      0        171        0   166
#> [323,]  67  10      1          9        0   200
#> [324,]  67   6      1          4        0   130
#> [325,]  69   0      0          0        1   148
#> [326,]  68   7      1          0        1   150
#> [327,]  69   3      1          2        0   151
#> [328,]  67  14      1         13        0   130
#> [329,]  71   7      0          7        0   230
#> [330,]  66   2      0          2        1   228
#> [331,]  71   6      0         45        1   158
#> [332,]  69   5      0          5        1   142
#> [333,]  71   3      0        103        0   133
#> [334,]  69   3      0          3        1   130
#> [335,]  67   1      0         36        1   104
#> [336,]  67   5      0          5        0   130
#> [337,]  72   3      1          0        1   132
#> [338,]  72   7      0          7        1   110
#> [339,]  69   8      1          7        1   108
#> [340,]  67  14      0        172        1   140
#> [341,]  69  11      1          0        1   120
#> [342,]  66   2      0        180        0   130
#> [343,]  67   7      1          4        0   122
#> [344,]  69   4      1          3        0   132
#> [345,]  68   2      0          7        1   130
#> [346,]  70   9      0        180        1   142
#> [347,]  72   5      1          4        0   170
#> [348,]  67  22      1          1        1   140
#> [349,]  68   3      0         19        0   135
#> [350,]  69   1      0          1        1   110
#> [351,]  67   1      0          1        1    60
#> [352,]  67   4      0         60        1   136
#> [353,]  69   5      0         76        0   120
#> [354,]  67   8      1          0        1   130
#> [355,]  68  10      1          8        1   160
#> [356,]  66  24      1         13        0   130
#> [357,]  70  35      1          0        1   105
#> [358,]  72  30      1          0        1   145
#> [359,]  70   7      0          7        0   102
#> [360,]  71   6      0          9        0   120
#> [361,]  69  10      1          6        1   120
#> [362,]  72  19      1          8        0   120
#> [363,]  72  12      1         10        0   170
#> [364,]  67   8      0        180        1   170
#> [365,]  67   9      0        180        0   158
#> [366,]  73  13      0        152        1   130
#> [367,]  70   5      0        180        0   150
#> [368,]  72   2      0          2        1   100
#> [369,]  67   4      1          1        0   134
#> [370,]  72   6      1          5        0   115
#> [371,]  71   1      0        173        1   188
#> [372,]  68  23      0        180        1   220
#> [373,]  70   3      0        180        0   121
#> [374,]  69   3      0        180        0   220
#> [375,]  68   4      1          3        0   210
#> [376,]  72   5      0         28        0   120
#> [377,]  69  16      1         10        1   140
#> [378,]  68   7      0        180        1   130
#> [379,]  72  16      1          1        1   130
#> [380,]  70   4      0        180        0   180
#> [381,]  69   1      1          0        0   155
#> [382,]  73   6      1          0        1   270
#> [383,]  72   8      1          1        1   150
#> [384,]  73   7      0          7        1   140
#> [385,]  68  15      1         13        1   130
#> [386,]  70   3      0          3        1   159
#> [387,]  72   6      0        180        1   130
#> [388,]  73   0      0        180        1   161
#> [389,]  74   8      1          0        1    85
#> [390,]  73   4      0        180        1   154
#> [391,]  69   2      1          0        1   110
#> [392,]  71   3      1          1        0   150
#> [393,]  71  15      1         11        0   165
#> [394,]  74  20      0         20        1   180
#> [395,]  68   9      0        180        1   120
#> [396,]  71  20      1         10        0   140
#> [397,]  74   0      1          0        1    90
#> [398,]  73   3      1          0        1   136
#> [399,]  70   5      1          0        1   190
#> [400,]  71  17      1         11        0   160
#> [401,]  71   8      1          7        0   149
#> [402,]  71   3      1          2        1   190
#> [403,]  73  10      1          8        0   106
#> [404,]  70  26      1         11        1   120
#> [405,]  74   4      0          4        0   120
#> [406,]  72   5      1          3        1   160
#> [407,]  70   3      0        180        1   154
#> [408,]  73   6      0        180        0   110
#> [409,]  69   2      1          1        1    80
#> [410,]  71  14      1         13        1   170
#> [411,]  72  15      1         13        0   156
#> [412,]  75   1      0          1        0   133
#> [413,]  75   2      1          1        0   145
#> [414,]  73  10      1          9        1   146
#> [415,]  72  10      1          9        1   160
#> [416,]  71   2      0         10        1   112
#> [417,]  73   1      0          1        1    80
#> [418,]  75   9      1          7        0   140
#> [419,]  75  13      1          1        1   130
#> [420,]  71  11      1          8        0   110
#> [421,]  71   4      0          4        0   134
#> [422,]  73  10      1          8        0   120
#> [423,]  70   7      1          4        0   184
#> [424,]  72   1      1          1        0   168
#> [425,]  72   7      0         57        1   145
#> [426,]  73  10      0        180        0   162
#> [427,]  72  11      0         11        1   140
#> [428,]  76  25      1         12        1   170
#> [429,]  73  12      1         12        1   140
#> [430,]  72   2      0        180        0   120
#> [431,]  71   3      1          0        0   144
#> [432,]  74  34      1          8        1   233
#> [433,]  72   5      0        180        0   154
#> [434,]  76   5      0          5        1   130
#> [435,]  77  11      0         11        1   150
#> [436,]  72   7      1          2        0   142
#> [437,]  71  16      0        180        0   140
#> [438,]  74   3      0          3        1   128
#> [439,]  76   1      0        180        0   114
#> [440,]  74   2      1          1        0   140
#> [441,]  76   8      1          0        1   141
#> [442,]  75  23      1         14        1   110
#> [443,]  74   2      0        180        0   190
#> [444,]  72   4      0         85        1   120
#> [445,]  72   4      1          3        0   160
#> [446,]  76  17      1          0        1   200
#> [447,]  73   4      1          3        1   125
#> [448,]  76  13      1         10        0   110
#> [449,]  75   4      1          0        1   122
#> [450,]  75   7      0          7        0   190
#> [451,]  75   0      0          0        1   130
#> [452,]  73  13      1         11        0   195
#> [453,]  74   8      1          0        1   105
#> [454,]  76  13      1          8        1   148
#> [455,]  75   4      1          2        1   188
#> [456,]  74   6      0        180        0   160
#> [457,]  76   4      0          4        1   155
#> [458,]  75   1      0          1        1   125
#> [459,]  72   5      0        180        0   120
#> [460,]  78  12      1         11        1   160
#> [461,]  76   5      0        180        0   185
#> [462,]  74  10      1          0        1   135
#> [463,]  74   8      1          8        1   170
#> [464,]  75   9      0        180        1   140
#> [465,]  73  33      1         12        1   175
#> [466,]  77   5      1          0        0   123
#> [467,]  73  10      1          9        0   146
#> [468,]  77   1      1          0        1    90
#> [469,]  76  12      1         11        1   120
#> [470,]  73   7      1          0        0   174
#> [471,]  75   3      1          1        1   171
#> [472,]  75   6      0        180        0   150
#> [473,]  79  10      1          8        0   190
#> [474,]  74   2      1          0        1   130
#> [475,]  78  18      0         18        1   144
#> [476,]  77   3      0        180        0   110
#> [477,]  76  29      0         47        0    90
#> [478,]  73   8      1          1        1   162
#> [479,]  73  11      1          2        1   110
#> [480,]  74   2      0        180        0   100
#> [481,]  78   7      0          7        1   133
#> [482,]  74  15      0        180        1   172
#> [483,]  78   8      1          6        1   110
#> [484,]  74   7      0          7        0   161
#> [485,]  76  13      1          1        1   170
#> [486,]  80  10      1          6        1   147
#> [487,]  78   0      0        180        1   212
#> [488,]  75  12      1          1        1   120
#> [489,]  74  10      1          8        0   135
#> [490,]  76   1      0          1        1    83
#> [491,]  79   4      0         80        0   145
#> [492,]  78  12      1          9        0   150
#> [493,]  78   2      1          1        0   130
#> [494,]  75   4      1          0        0   212
#> [495,]  76  11      1          0        0   120
#> [496,]  75   3      0          3        0     0
#> [497,]  76   7      0         29        1   150
#> [498,]  79   8      0         32        1   120
#> [499,]  80   9      0         23        1   128
#> [500,]  80   6      0          6        1   150
#> [501,]  76   3      1          0        1   140
#> [502,]  78  11      1          1        1   140
#> [503,]  79   2      1          0        1   121
#> [504,]  78  14      1          0        1   140
#> [505,]  78  11      1          8        1   118
#> [506,]  76   4      0          4        1   160
#> [507,]  79   4      0          4        1   125
#> [508,]  76  10      1          8        0   180
#> [509,]  76  12      1         10        1   127
#> [510,]  80   3      1          0        1   120
#> [511,]  75   2      1          1        1   204
#> [512,]  78  11      0        180        1   135
#> [513,]  76   1      0          1        1   140
#> [514,]  78   7      1          0        1   110
#> [515,]  79   3      0          3        0   120
#> [516,]  77   7      0        180        1   170
#> [517,]  77   6      0          6        1   144
#> [518,]  79   4      1          0        1   120
#> [519,]  81   1      0        180        0   120
#> [520,]  80  15      1         12        1   150
#> [521,]  77   9      1          4        0   141
#> [522,]  82   5      0          8        1   120
#> [523,]  80  40      1          0        1   138
#> [524,]  76   7      0        161        0   151
#> [525,]  80  15      1          0        1    90
#> [526,]  81   4      1          2        1   126
#> [527,]  79  28      0        164        0   100
#> [528,]  80   6      0        173        1   160
#> [529,]  79   1      0         37        1   140
#> [530,]  81   3      0        180        0   184
#> [531,]  81   2      0        175        0   172
#> [532,]  78   7      0          7        1   147
#> [533,]  78  15      0         15        0   165
#> [534,]  80   5      1          1        1   108
#> [535,]  79   3      0          3        1   101
#> [536,]  78  26      1          5        0   194
#> [537,]  76   1      0        166        0   131
#> [538,]  78   3      1          1        1   152
#> [539,]  77  10      1          8        1   130
#> [540,]  82   3      1          1        1   144
#> [541,]  80   2      1          1        0   168
#> [542,]  79   6      0          6        0   152
#> [543,]  80   6      1          0        1   119
#> [544,]  78   2      0        180        0   148
#> [545,]  80   5      0          5        1   130
#> [546,]  82   1      1          0        1    82
#> [547,]  79  10      0        180        1   150
#> [548,]  77   4      0        180        1    98
#> [549,]  79   1      0        125        0   193
#> [550,]  82  21      1          2        0   155
#> [551,]  84  22      1         10        0   180
#> [552,]  79   4      0          4        1   121
#> [553,]  80   6      0          6        1   110
#> [554,]  83   9      1          5        1   170
#> [555,]  82   5      0        180        0   110
#> [556,]  83   4      0        103        0    97
#> [557,]  80  11      1          8        0   170
#> [558,]  79   4      0          4        1   183
#> [559,]  78   9      1          4        1   120
#> [560,]  79   1      0        180        1   170
#> [561,]  81  15      0        180        1   140
#> [562,]  80   7      1          0        1   146
#> [563,]  81  16      0         16        1   110
#> [564,]  80   6      1          0        1   150
#> [565,]  81   8      0        180        0   146
#> [566,]  79   7      0        177        0   197
#> [567,]  85   4      0        180        0    90
#> [568,]  81   2      1          1        0   198
#> [569,]  82   6      0        128        1   100
#> [570,]  84   4      0        167        0   198
#> [571,]  82  23      1          0        0   110
#> [572,]  84   5      0        180        1   203
#> [573,]  81   1      0          1        1   150
#> [574,]  84   1      0         38        1   205
#> [575,]  83   3      0        180        0   174
#> [576,]  85   3      1          2        1   160
#> [577,]  84   4      0         89        1   129
#> [578,]  80   2      1          0        1   130
#> [579,]  80   6      0         71        1   189
#> [580,]  83   1      0          1        1   100
#> [581,]  82  19      0         19        0   120
#> [582,]  80  30      1         13        0   220
#> [583,]  83   9      0        180        0   198
#> [584,]  79  14      1          0        0   110
#> [585,]  83   3      0        114        0    98
#> [586,]  81  14      1         12        1   128
#> [587,]  83   2      0        154        0   130
#> [588,]  85   9      1          6        1   160
#> [589,]  81   4      0          4        0   175
#> [590,]  84  15      1         13        1   110
#> [591,]  82   5      1          0        1   146
#> [592,]  81   4      0          4        0   160
#> [593,]  86  12      0        180        1   120
#> [594,]  83  12      1          2        1   170
#> [595,]  81  19      1         14        0   120
#> [596,]  82   3      1          2        0   130
#> [597,]  80   2      0         88        0   135
#> [598,]  83   7      0        126        0   135
#> [599,]  81  16      1          9        0   180
#> [600,]  84   6      0        165        0   145
#> [601,]  86   3      0          3        1   140
#> [602,]  84   3      0        180        1   120
#> [603,]  85   3      0          3        1   118
#> [604,]  81   4      0        180        0   160
#> [605,]  83   9      0        180        1   149
#> [606,]  87   2      0          5        1   137
#> [607,]  86  12      1          0        1   132
#> [608,]  86   6      1          0        1   140
#> [609,]  84   3      1          2        0   125
#> [610,]  83  10      1          0        1   190
#> [611,]  88  14      1          3        1   130
#> [612,]  84   3      0          3        1   121
#> [613,]  83  13      1         12        0   170
#> [614,]  84   9      0         92        1   110
#> [615,]  84   3      0        180        1   170
#> [616,]  86   4      0         38        1   122
#> [617,]  86  13      0        177        0   163
#> [618,]  83  20      1          3        1   150
#> [619,]  88   4      0          4        1   115
#> [620,]  85  22      0         22        1   184
#> [621,]  87   2      0        180        1   130
#> [622,]  86   6      0         46        0   173
#> [623,]  88   3      0        115        0   110
#> [624,]  83   3      0          3        1   130
#> [625,]  87   8      0          8        1   157
#> [626,]  88   4      0          4        0    86
#> [627,]  89   4      0          4        1   153
#> [628,]  89   5      0        119        1   140
#> [629,]  84   8      0        180        1   119
#> [630,]  85   8      0          8        1   136
#> [631,]  84   2      0        110        1   174
#> [632,]  84   0      0        180        1   136
#> [633,]  90  14      0         14        1   100
#> [634,]  88   1      0          1        0   135
#> [635,]  91   8      0          8        0   100
#> [636,]  87   2      0        180        0   160
#> [637,]  87   6      1          0        0   125
#> [638,]  91  10      0        145        0   135
#> [639,]  86   3      1          0        1    80
#> [640,]  88   7      0         24        0   119
#> [641,]  86  10      0        180        1   137
#> [642,]  86   9      1          7        0   130
#> [643,]  87  43      0        178        1   130
#> [644,]  87   5      0         36        1   150
#> [645,]  90   5      1          0        1   125
#> [646,]  88   3      1          2        0   159
#> [647,]  89   3      1          1        1   160
#> [648,]  92   1      0          1        1   167
#> [649,]  91   3      0         33        1   137
#> [650,]  88   5      0        158        0   100
#> [651,]  87   7      0         74        1   105
#> [652,]  89  12      1          0        1   130
#> [653,]  89   2      0        168        0   118
#> [654,]  91   5      0        169        1   176
#> [655,]  92   7      0          7        1   110
#> [656,]  91   0      0          0        0     0
#> [657,]  89  14      0        180        1    84
#> [658,]  91   4      1          0        1   120
#> [659,]  93   8      0        179        1   110
#> [660,]  94   8      0          8        1   142
#> [661,]  92   4      0         76        1   149
#> [662,]  91   1      0        180        0   158
#> [663,]  90  16      0         16        1   106
#> [664,]  96   3      0         12        1    97
#> [665,]  95   8      1          5        1   150
#> [666,]  91  12      0         53        1   212
#> [667,]  91   7      0          7        0   135
#> [668,]  92   2      0          2        0   112
#> [669,]  96   3      1          0        1   104
#> [670,]  96  15      1          0        1   140
#> 
#> $y
#>   [1] 180.0+   5.0+   5.0+ 180.0+   2.0+ 180.0+ 180.0+ 115.0   12.0  180.0+
#>  [11] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 180.0+   2.0+ 180.0+ 180.0+
#>  [21] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 155.0+ 180.0+ 180.0+ 180.0+
#>  [31] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 150.0  180.0+
#>  [41] 180.0+ 180.0+   6.0+ 180.0+ 180.0+ 180.0+  73.0    5.0+ 161.0+ 180.0+
#>  [51]   5.0+ 180.0+ 180.0+ 180.0+ 177.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [61] 172.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0    1.0  179.0+ 180.0+ 179.0+
#>  [71] 180.0+   4.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [81] 180.0+  36.0   88.0+   4.0+ 180.0+ 180.0+ 180.0+  77.0  180.0+ 180.0+
#>  [91]   5.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0+  85.0  180.0+
#> [101] 180.0+ 166.0+  99.0  180.0+  16.0+ 180.0+ 180.0+ 180.0+  13.0+ 171.0+
#> [111] 180.0+ 174.0+   6.0+   1.0    9.0+ 180.0+ 180.0+   7.0+ 180.0+ 180.0+
#> [121] 180.0+ 180.0+ 180.0+ 180.0+  16.0   15.0+ 180.0+ 180.0+  12.0+   8.0 
#> [131] 180.0+ 180.0+ 180.0+ 180.0+   1.0  165.0  180.0+ 180.0+ 180.0+ 180.0+
#> [141] 180.0+ 180.0+ 180.0+   0.5  180.0+ 180.0+ 180.0+   5.0+ 180.0+   4.0+
#> [151] 147.0+ 180.0+   5.0+ 180.0+ 180.0+   1.0  180.0+ 180.0+   2.0+ 180.0+
#> [161] 180.0+   2.0  180.0+  64.0  180.0+ 180.0+ 180.0+   0.5  161.0+ 171.0+
#> [171] 180.0+ 180.0+   1.0  180.0+ 180.0+  10.0+ 180.0+ 180.0+  45.0    3.0+
#> [181]  19.0  180.0+   9.0+ 180.0+ 172.0+ 172.0+   8.0  180.0+ 180.0+   1.0+
#> [191] 180.0+  77.0   13.0+ 180.0+ 170.0   94.0  180.0+   7.0    7.0+ 180.0+
#> [201] 180.0+ 180.0+   3.0+  13.0+   5.0  180.0+ 180.0+   1.0  180.0+ 180.0+
#> [211] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+ 180.0+ 180.0+   9.0 
#> [221] 180.0+  84.0    7.0+ 180.0+ 180.0+ 180.0+   1.0    1.0  180.0+   3.0+
#> [231]   6.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  14.0+  36.0  180.0+   3.0+
#> [241]  88.0  180.0+ 180.0+ 180.0+   0.5  180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [251]  12.0  180.0+   9.0    3.0  180.0+ 180.0+  11.0+   2.0+ 180.0+ 180.0+
#> [261] 180.0+ 180.0+   3.0+   2.0+ 103.0   15.0  180.0+  13.0  179.0+ 166.0+
#> [271]   3.0    0.5+   3.0+ 180.0+ 180.0+ 180.0+   8.0   16.0  180.0+ 180.0+
#> [281] 180.0+ 123.0+   1.0+ 180.0+  80.0   15.0+ 180.0+ 180.0+ 180.0+ 174.0+
#> [291] 180.0+ 180.0+   8.0+   3.0  175.0  180.0+  10.0  180.0+ 180.0+   6.0 
#> [301]  12.0   99.0  180.0+ 179.0+ 180.0+ 180.0+  18.0    7.0+ 180.0+  93.0 
#> [311]  21.0+ 180.0+  18.0+ 101.0    4.0    5.0    7.0+ 180.0+   1.0  180.0+
#> [321] 180.0+ 171.0  174.0+   6.0    0.5  180.0+ 180.0+ 180.0+   7.0+   2.0 
#> [331]  45.0    5.0+ 103.0    3.0+  36.0    5.0+ 180.0+   7.0    8.0+ 172.0+
#> [341] 180.0+ 180.0+   7.0  180.0+   7.0  180.0+ 180.0+  51.0   19.0    1.0 
#> [351]   1.0   60.0   76.0  180.0+  10.0+ 180.0+ 180.0+ 162.0    7.0+   9.0 
#> [361] 180.0+ 180.0+  12.0  180.0+ 180.0+ 152.0  180.0+   2.0   76.0  180.0+
#> [371] 173.0+ 180.0+ 180.0+ 180.0+ 180.0+  28.0   16.0+ 180.0+  16.0+ 180.0+
#> [381] 180.0+   6.0  180.0+   7.0+  15.0    3.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [391]   2.0    3.0+ 180.0+  20.0  180.0+  20.0    0.5  180.0+ 180.0+ 180.0+
#> [401]   8.0    3.0   87.0  180.0+   4.0+ 180.0+ 180.0+ 180.0+   2.0   14.0+
#> [411] 180.0+   1.0  180.0+ 180.0+ 159.0   10.0    1.0  180.0+  13.0  180.0+
#> [421]   4.0+  10.0  104.0+   1.0   57.0  180.0+  11.0  180.0+  12.0  180.0+
#> [431] 180.0+  34.0  180.0+   5.0   11.0+   7.0  180.0+   3.0  180.0+ 180.0+
#> [441] 180.0+ 180.0+ 180.0+  85.0  180.0+  17.0+ 180.0+ 174.0+   4.0    7.0 
#> [451]   0.5  180.0+ 180.0+ 180.0+  46.0  180.0+   4.0    1.0  180.0+  12.0 
#> [461] 180.0+ 180.0+   8.0  180.0+  33.0    5.0  180.0+   1.0   12.0    7.0+
#> [471]   3.0  180.0+ 180.0+ 176.0+  18.0  180.0+  47.0  180.0+  11.0  180.0+
#> [481]   7.0  180.0+   8.0+   7.0  180.0+  10.0  180.0+  12.0  180.0+   1.0 
#> [491]  80.0  180.0+ 180.0+   4.0+  11.0    3.0   29.0   32.0   23.0    6.0 
#> [501]   3.0+ 180.0+ 180.0+ 180.0+  11.0    4.0    4.0   10.0+ 180.0+   3.0+
#> [511]   2.0+ 180.0+   1.0   43.0    3.0  180.0+   6.0  138.0  180.0+ 180.0+
#> [521]  71.0    8.0   40.0  161.0  180.0+  93.0  164.0  173.0   37.0  180.0+
#> [531] 175.0+   7.0+  15.0+   5.0+   3.0  171.0+ 166.0+   3.0+  10.0  180.0+
#> [541]  10.0    6.0+   6.0  180.0+   5.0    1.0  180.0+ 180.0+ 125.0  180.0+
#> [551] 180.0+   4.0    6.0    9.0+ 180.0+ 103.0  169.0    4.0  180.0+ 180.0+
#> [561] 180.0+   7.0+  16.0  180.0+ 180.0+ 177.0+ 180.0+ 180.0+ 128.0  167.0 
#> [571]  62.0  180.0+   1.0   38.0  180.0+ 180.0+  89.0  180.0+  71.0    1.0 
#> [581]  19.0   30.0  180.0+ 180.0+ 114.0  180.0+ 154.0  180.0+   4.0+ 180.0+
#> [591]   5.0+   4.0+ 180.0+  77.0  180.0+   3.0   88.0  126.0  180.0+ 165.0 
#> [601]   3.0  180.0+   3.0+ 180.0+ 180.0+   5.0  180.0+   6.0  180.0+ 180.0+
#> [611]  14.0    3.0   13.0   92.0  180.0+  38.0  177.0   20.0    4.0   22.0 
#> [621] 180.0+  46.0  115.0    3.0+   8.0+   4.0    4.0  119.0  180.0+   8.0 
#> [631] 110.0  180.0+  14.0    1.0+   8.0  180.0+  25.0  145.0    3.0   24.0 
#> [641] 180.0+ 180.0+ 178.0+  36.0   89.0   75.0    3.0+   1.0   33.0  158.0 
#> [651]  74.0  180.0+ 168.0  169.0    7.0    0.5  180.0+   4.0  179.0+   8.0+
#> [661]  76.0  180.0+  16.0   12.0    8.0   53.0    7.0+   2.0    3.0   15.0+
#> 
#> $weights
#> NULL
#> 
#> $offset
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Warning: Multiple lambdas have been fit. Lambda will be set to 0.01 (see parameter 's').

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8411329