GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::glmnet() from package glmnet.

Details

This learner returns two prediction types:

lp: a vector of linear predictors (relative risk scores), one per observation. Calculated using glmnet::predict.coxnet().

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.coxnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used.

Caution: This learner is different to learners calling glmnet::cv.glmnet() in that it does not use the internal optimization of parameter lambda. Instead, lambda needs to be tuned by the user (e.g., via mlr3tuning). When lambda is tuned, the glmnet will be trained for each tuning iteration. While fitting the whole path of lambdas would be more efficient, as is done by default in glmnet::glmnet(), tuning/selecting the parameter at prediction time (using parameter s) is currently not supported in mlr3 (at least not in efficient manner). Tuning the s parameter is, therefore, currently discouraged.

When the data are i.i.d. and efficiency is key, we recommend using the respective auto-tuning counterpart in mlr_learners_surv.cv_glmnet(). However, in some situations this is not applicable, usually when data are imbalanced or not i.i.d. (longitudinal, time-series) and tuning requires custom resampling strategies (blocked design, stratification).

Custom mlr3 parameters

family is set to "cox" and cannot be changed.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exact	logical	FALSE	TRUE, FALSE	-

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

newoffset	untyped	-		-

nlambda	integer	100		\([1, \infty)\)

offset	untyped	NULL		-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	0.01		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.logistic	character	Newton	Newton, modified.Newton	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvGlmnet

Methods

Public methods

LearnerSurvGlmnet$new()

LearnerSurvGlmnet$selected_features()

LearnerSurvGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = mlr3::lrn("surv.glmnet")
print(learner)
#> <LearnerSurvGlmnet:surv.glmnet>: Regularized Generalized Linear Model
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3proba, mlr3extralearners, glmnet
#> * Predict Types:  [crank], distr, lp
#> * Feature Types: logical, integer, numeric
#> * Properties: selected_features, weights

# Define a Task
task = mlr3::tsk("grace")

# Create train and test set
ids = mlr3::partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#>    Df  %Dev   Lambda
#> 1   0  0.00 0.177800
#> 2   2  0.81 0.162000
#> 3   2  1.76 0.147600
#> 4   2  2.58 0.134500
#> 5   3  4.58 0.122500
#> 6   3  7.13 0.111600
#> 7   3  9.04 0.101700
#> 8   3 10.52 0.092690
#> 9   3 11.71 0.084460
#> 10  3 12.67 0.076960
#> 11  4 13.50 0.070120
#> 12  4 14.21 0.063890
#> 13  4 14.80 0.058210
#> 14  5 15.30 0.053040
#> 15  6 15.77 0.048330
#> 16  6 16.18 0.044040
#> 17  6 16.52 0.040120
#> 18  6 16.80 0.036560
#> 19  6 17.05 0.033310
#> 20  6 17.26 0.030350
#> 21  6 17.43 0.027660
#> 22  6 17.58 0.025200
#> 23  6 17.70 0.022960
#> 24  6 17.81 0.020920
#> 25  6 17.90 0.019060
#> 26  6 17.97 0.017370
#> 27  6 18.03 0.015830
#> 28  6 18.09 0.014420
#> 29  6 18.13 0.013140
#> 30  6 18.17 0.011970
#> 31  6 18.20 0.010910
#> 32  6 18.23 0.009939
#> 33  6 18.25 0.009056
#> 34  6 18.27 0.008252
#> 35  6 18.28 0.007519
#> 36  6 18.30 0.006851
#> 37  6 18.31 0.006242
#> 38  6 18.32 0.005688
#> 39  6 18.32 0.005182
#> 40  6 18.33 0.004722
#> 41  6 18.34 0.004302
#> 42  6 18.34 0.003920
#> 43  6 18.34 0.003572
#> 44  6 18.35 0.003255
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  28   9      0        180        1   107
#>   [2,]  32   5      1          0        1   121
#>   [3,]  33   2      0          2        0   150
#>   [4,]  35  10      1          9        0   106
#>   [5,]  34   5      0          5        0   120
#>   [6,]  35   2      0        180        0   121
#>   [7,]  35   2      1          1        1   112
#>   [8,]  37   9      0        180        1   151
#>   [9,]  38  13      1          0        1   161
#>  [10,]  35   0      0        180        1   119
#>  [11,]  36   5      1          0        1   115
#>  [12,]  38  16      1         10        0   160
#>  [13,]  38  12      1         11        1    92
#>  [14,]  40  12      1          9        0   153
#>  [15,]  42   3      1          1        1   130
#>  [16,]  40   2      1          1        1   148
#>  [17,]  42   2      0        180        1   100
#>  [18,]  42   2      0          2        0   140
#>  [19,]  40  11      1         10        1   120
#>  [20,]  42   2      0        180        0   100
#>  [21,]  43   3      1          0        1   100
#>  [22,]  41   2      1          1        0   166
#>  [23,]  40   1      1          0        1   145
#>  [24,]  43   4      1          0        1   130
#>  [25,]  42   4      0        180        0   162
#>  [26,]  42  15      1         13        1   125
#>  [27,]  42   2      0        180        1   124
#>  [28,]  44   5      1          1        0   170
#>  [29,]  45   3      0        180        1   154
#>  [30,]  44   3      0        180        0   141
#>  [31,]  45   6      0        180        1   170
#>  [32,]  43   2      0        180        1   140
#>  [33,]  45   2      0        180        1   140
#>  [34,]  46  15      0        180        0   120
#>  [35,]  46   2      1          1        0   126
#>  [36,]  47   4      1          3        0   118
#>  [37,]  45   3      0        150        0   130
#>  [38,]  43  29      0        180        1   180
#>  [39,]  45   4      1          0        0   124
#>  [40,]  47   6      1          0        1   116
#>  [41,]  47   4      1          3        1   160
#>  [42,]  45   8      1          0        1   117
#>  [43,]  49   5      0         73        1   136
#>  [44,]  46   6      1          0        1   100
#>  [45,]  47   2      0        180        0   108
#>  [46,]  44   9      1          8        1   135
#>  [47,]  46   4      0        180        1   121
#>  [48,]  46  15      0        180        1   120
#>  [49,]  47   3      1          1        1   120
#>  [50,]  48   3      0        180        0   154
#>  [51,]  48  12      1         11        0   200
#>  [52,]  46   3      1          0        1   119
#>  [53,]  47  10      0         10        1   140
#>  [54,]  50   1      1          0        1   129
#>  [55,]  47   7      0        180        0   145
#>  [56,]  50   4      1          1        0   125
#>  [57,]  50   6      1          2        1   140
#>  [58,]  49   7      1          7        1   110
#>  [59,]  46   3      1          1        1   140
#>  [60,]  49   2      0          2        0   105
#>  [61,]  51   1      0          1        1   145
#>  [62,]  49  15      1         11        1   160
#>  [63,]  47   2      0        180        0   150
#>  [64,]  49  23      0        179        1   112
#>  [65,]  46   6      1          0        1   156
#>  [66,]  50   7      1          0        1    92
#>  [67,]  50   4      0          4        1   100
#>  [68,]  51   3      1          2        0   113
#>  [69,]  47   8      0        180        0   160
#>  [70,]  52   2      0        180        0   155
#>  [71,]  46   3      0        180        1   120
#>  [72,]  46   1      1          1        0   145
#>  [73,]  50   4      1          1        0   150
#>  [74,]  48   7      1          0        1   110
#>  [75,]  53   8      0         36        1   160
#>  [76,]  48  17      1         10        0   111
#>  [77,]  47   2      1          1        0   110
#>  [78,]  52   4      1          4        0   152
#>  [79,]  53   5      0        180        1   140
#>  [80,]  54  17      1         12        1   102
#>  [81,]  53   7      1          0        0   199
#>  [82,]  51   3      1          1        0   140
#>  [83,]  50   2      0          5        1   106
#>  [84,]  50  10      1          6        0   122
#>  [85,]  50  14      1         13        0   170
#>  [86,]  53   8      1          7        0   160
#>  [87,]  48   3      1          2        0   150
#>  [88,]  49   5      1          2        1   150
#>  [89,]  53   4      0          4        0   140
#>  [90,]  48   6      0        180        0   160
#>  [91,]  51  13      0         99        1   160
#>  [92,]  54   9      1          0        1   138
#>  [93,]  49  16      0         16        0   125
#>  [94,]  55   3      1          1        0   150
#>  [95,]  52   7      1          2        0   154
#>  [96,]  55   6      1          2        1   114
#>  [97,]  52   4      0        180        1   180
#>  [98,]  52   4      0        180        0   183
#>  [99,]  49   6      1          0        1   130
#> [100,]  49   1      0          1        1   110
#> [101,]  50   7      1          1        0   156
#> [102,]  53   9      0          9        1    95
#> [103,]  53   8      1          0        1   130
#> [104,]  50   7      1          0        1   127
#> [105,]  52   5      0        175        1   117
#> [106,]  55   1      0        180        0   127
#> [107,]  56   3      0        180        1   193
#> [108,]  56   2      0        180        0   132
#> [109,]  55   5      1          4        1   120
#> [110,]  53  18      1          9        1   150
#> [111,]  54   3      0        180        1   180
#> [112,]  55   6      0        180        0   170
#> [113,]  53  10      1          9        0   172
#> [114,]  52  16      1         14        0   170
#> [115,]  53  15      0         15        1    90
#> [116,]  53   4      0        180        1   150
#> [117,]  55   6      0        180        1   100
#> [118,]  55   6      1          5        1   138
#> [119,]  55   2      0        134        1   140
#> [120,]  56   3      0          8        1   139
#> [121,]  55   1      0          2        0   130
#> [122,]  57   3      0          3        0   120
#> [123,]  54   7      1          2        0   129
#> [124,]  54   2      1          1        0   135
#> [125,]  52   9      1          3        0   170
#> [126,]  57   5      1          3        1   138
#> [127,]  57   1      0        180        1   156
#> [128,]  57   1      0          1        1   100
#> [129,]  52   2      0        180        0   140
#> [130,]  55  11      1          7        0   104
#> [131,]  52  15      1         14        0   130
#> [132,]  56  14      1         11        0   130
#> [133,]  57  10      0        180        1   170
#> [134,]  58   8      0          8        1   130
#> [135,]  54   5      0        180        1   108
#> [136,]  55   3      1          1        1   156
#> [137,]  53  21      1         13        1   130
#> [138,]  59   3      1          1        0   172
#> [139,]  57   4      0        180        1   119
#> [140,]  58   6      1          0        1    90
#> [141,]  53  15      1         10        1   130
#> [142,]  54  17      1          8        1   227
#> [143,]  55   9      1          2        1   147
#> [144,]  55  13      0        166        1   140
#> [145,]  56   5      0          5        1   150
#> [146,]  57   4      1          2        1   185
#> [147,]  53   7      1          0        1   120
#> [148,]  55   5      0          5        1   131
#> [149,]  54   7      1          0        1   141
#> [150,]  56   4      0          4        0   164
#> [151,]  59  15      1         10        0   140
#> [152,]  58   9      1          0        1   180
#> [153,]  55   5      1          0        0   140
#> [154,]  55   2      0          2        0   106
#> [155,]  59   9      1          1        1   125
#> [156,]  57   1      0        180        0   148
#> [157,]  58   4      1          0        1   160
#> [158,]  57   2      0          2        1   120
#> [159,]  60   5      1          1        0   138
#> [160,]  57   5      0        180        1   130
#> [161,]  58  11      1          9        1   124
#> [162,]  55   5      1          0        1   160
#> [163,]  59   6      1          0        1   140
#> [164,]  59   5      0        180        1   155
#> [165,]  59   4      1          0        1   152
#> [166,]  58  26      1          0        1   189
#> [167,]  61   9      0          9        1   160
#> [168,]  58   4      1          3        0   120
#> [169,]  60   0      1          0        1    80
#> [170,]  58   8      0        161        1   140
#> [171,]  58  14      1          6        0   190
#> [172,]  61   4      1          3        0   151
#> [173,]  61   9      1          8        0   150
#> [174,]  58   1      0          1        1   100
#> [175,]  61  20      1         13        0   130
#> [176,]  57  13      1         10        0   110
#> [177,]  57   2      1          0        1   116
#> [178,]  58  10      0         10        1   150
#> [179,]  57  11      0        180        1   150
#> [180,]  61   3      0         17        0   143
#> [181,]  56  14      0         45        0   130
#> [182,]  57   3      1          2        0   120
#> [183,]  56  13      1          6        1   158
#> [184,]  56  18      1         11        1   165
#> [185,]  60  12      1          0        1   114
#> [186,]  55   9      1          7        1   135
#> [187,]  56   8      1          8        0   120
#> [188,]  61  13      1         12        1   130
#> [189,]  59  11      1          8        1   190
#> [190,]  57   1      0          1        0   126
#> [191,]  59   5      1          2        0   182
#> [192,]  58   5      1          1        1   135
#> [193,]  61   8      0         77        0   120
#> [194,]  61  13      0         13        0   210
#> [195,]  62  10      1          0        1   153
#> [196,]  57   3      1          0        0   100
#> [197,]  61  28      1          7        0   133
#> [198,]  58   8      1          3        1   150
#> [199,]  57   7      0        169        0   180
#> [200,]  61   7      0          7        1   150
#> [201,]  60   7      0          7        0   147
#> [202,]  62   4      1          0        0   160
#> [203,]  60  17      1          8        1   140
#> [204,]  62   4      1          3        0   173
#> [205,]  59   1      0        180        0   155
#> [206,]  59  16      1          9        1   133
#> [207,]  61  13      0         13        0   120
#> [208,]  61   5      0          5        1   110
#> [209,]  57  18      1          9        1    93
#> [210,]  61   5      0          5        1   160
#> [211,]  58  11      1          9        0   179
#> [212,]  57   2      1          1        0   159
#> [213,]  62   1      1          0        1   172
#> [214,]  58   7      0        180        1   150
#> [215,]  63   3      1          1        0   180
#> [216,]  63   1      0        180        1   130
#> [217,]  63   4      0        180        1   190
#> [218,]  63  15      1         10        1   126
#> [219,]  63   4      1          1        0   155
#> [220,]  60  18      1         13        0   132
#> [221,]  59   8      0        180        1   140
#> [222,]  61   9      1          9        1   150
#> [223,]  58   9      1          9        0   110
#> [224,]  58   2      0        180        0   127
#> [225,]  59   4      0        180        0   196
#> [226,]  60   7      1          5        1   141
#> [227,]  60   7      0          7        0   140
#> [228,]  60   7      1          1        1    90
#> [229,]  65  13      0        180        1   100
#> [230,]  62   6      0        180        0   170
#> [231,]  59   4      0          4        0   149
#> [232,]  64  10      1          9        0   160
#> [233,]  59  10      0        180        1   130
#> [234,]  60   8      0         17        1   130
#> [235,]  61   6      1          1        1   117
#> [236,]  64  12      1         11        0   160
#> [237,]  66   1      1          0        1   120
#> [238,]  65  36      1         11        0   140
#> [239,]  63   4      1          3        0   162
#> [240,]  61  10      1          2        1   194
#> [241,]  63  12      1          9        0   114
#> [242,]  63   7      0        180        0   120
#> [243,]  65   8      1          0        0   168
#> [244,]  64   0      0          0        1    90
#> [245,]  64  21      1         10        0   190
#> [246,]  61  12      1         11        0   154
#> [247,]  64   9      0        180        0   150
#> [248,]  63  16      1          7        1   110
#> [249,]  64   7      0        180        1   120
#> [250,]  66   6      1          1        1   130
#> [251,]  62   3      1          1        1   199
#> [252,]  63   5      1          4        0   170
#> [253,]  63   2      1          1        0   180
#> [254,]  66  18      1          5        0   142
#> [255,]  66  16      1         11        1   169
#> [256,]  62   9      0        180        0   145
#> [257,]  61  14      1          5        0   140
#> [258,]  61  15      1         10        0   130
#> [259,]  63   9      1          8        1   160
#> [260,]  63   3      1          2        0   120
#> [261,]  63   2      1          0        0   140
#> [262,]  64  19      1          8        1   160
#> [263,]  65   8      1          0        1   140
#> [264,]  67   6      0        180        1   170
#> [265,]  64  13      1         12        1   150
#> [266,]  64   6      1          0        1   125
#> [267,]  66   7      1          0        1   115
#> [268,]  66  13      1          0        0   118
#> [269,]  65   3      0          3        0   105
#> [270,]  64   0      0          0        1   148
#> [271,]  67   4      1          3        0   130
#> [272,]  68   1      0        180        1   166
#> [273,]  64  10      1          9        1   110
#> [274,]  63   7      1          0        0   162
#> [275,]  67   8      1          1        1   130
#> [276,]  66  14      0        180        0   130
#> [277,]  68  18      0        180        1   260
#> [278,]  65  17      1         14        1   100
#> [279,]  63   8      1          1        1   162
#> [280,]  65  18      1          3        0   120
#> [281,]  63   1      1          0        1   155
#> [282,]  63  10      0         18        1   130
#> [283,]  67  11      0         11        0   150
#> [284,]  68  11      0        180        0   160
#> [285,]  65  15      1         12        1   150
#> [286,]  66  11      1          0        0   100
#> [287,]  69  12      0         15        1   140
#> [288,]  66  15      1         13        1   160
#> [289,]  63   2      0        180        0   150
#> [290,]  69  21      1         10        0   180
#> [291,]  66   9      1          8        0   130
#> [292,]  69   1      1          0        0   170
#> [293,]  67   1      0        180        1   160
#> [294,]  68  10      1         10        1   150
#> [295,]  65   1      1          0        0   133
#> [296,]  67   7      1          4        1   130
#> [297,]  63   2      1          0        0    99
#> [298,]  67   2      0        180        0   184
#> [299,]  65  10      1          1        1   148
#> [300,]  67  12      1         12        0   160
#> [301,]  69   6      0         99        1   140
#> [302,]  64   4      0        179        0   160
#> [303,]  66   4      0        180        1   130
#> [304,]  70  15      1         12        1   132
#> [305,]  64  11      0         11        0   125
#> [306,]  64   4      0        180        1   140
#> [307,]  64   0      1          0        1   118
#> [308,]  66   7      1          5        1   131
#> [309,]  68   4      1          0        1   160
#> [310,]  65  13      1         12        1   130
#> [311,]  69  17      1         10        0   140
#> [312,]  69   8      0         93        0   140
#> [313,]  65   1      0          1        1   120
#> [314,]  68  18      1          0        1   160
#> [315,]  68   4      0          4        1   190
#> [316,]  71   3      0          5        0   112
#> [317,]  70   7      1          0        1   190
#> [318,]  71  20      1          0        1   160
#> [319,]  67   2      0        180        0   128
#> [320,]  66   9      1          3        1   151
#> [321,]  66   1      1          1        1   165
#> [322,]  69   8      0        180        1   153
#> [323,]  70  14      0        171        0   166
#> [324,]  66   4      0        180        0   130
#> [325,]  67  10      1          9        0   200
#> [326,]  67   6      1          4        0   130
#> [327,]  68  18      1         14        1   170
#> [328,]  69   0      0          0        1   148
#> [329,]  65   2      0        180        0   130
#> [330,]  68   7      1          0        1   150
#> [331,]  67  14      1         13        0   130
#> [332,]  65  14      1         13        1   150
#> [333,]  66   2      0          2        1   228
#> [334,]  71   3      0        103        0   133
#> [335,]  69   3      0          3        1   130
#> [336,]  67   1      0         36        1   104
#> [337,]  67   5      0          5        0   130
#> [338,]  68   6      0        180        0   145
#> [339,]  72   7      0          7        1   110
#> [340,]  69   8      1          7        1   108
#> [341,]  67   3      0        180        0   110
#> [342,]  69  19      0        180        0   130
#> [343,]  68  18      0         18        1   100
#> [344,]  67  14      0        172        1   140
#> [345,]  69  11      1          0        1   120
#> [346,]  66   2      0        180        0   130
#> [347,]  67   7      1          4        0   122
#> [348,]  69   4      1          3        0   132
#> [349,]  68   2      0          7        1   130
#> [350,]  67  13      1          9        0   130
#> [351,]  72   5      1          4        0   170
#> [352,]  67  22      1          1        1   140
#> [353,]  68   3      0         19        0   135
#> [354,]  67   1      0          1        1    60
#> [355,]  67   4      0         60        1   136
#> [356,]  69   5      0         76        0   120
#> [357,]  67   8      1          0        1   130
#> [358,]  72  13      1         11        1   195
#> [359,]  66  24      1         13        0   130
#> [360,]  72  30      1          0        1   145
#> [361,]  70   7      0          7        0   102
#> [362,]  73  20      1          0        1   170
#> [363,]  71   6      0          9        0   120
#> [364,]  69  10      1          6        1   120
#> [365,]  70  11      0        180        1   210
#> [366,]  72  19      1          8        0   120
#> [367,]  67   8      0        180        1   170
#> [368,]  67   9      0        180        0   158
#> [369,]  73  13      0        152        1   130
#> [370,]  72   2      0          2        1   100
#> [371,]  71   1      0        173        1   188
#> [372,]  70   3      0        180        0   121
#> [373,]  69   3      0        180        0   220
#> [374,]  72   5      0         28        0   120
#> [375,]  71   5      0        180        0   191
#> [376,]  69  16      1         10        1   140
#> [377,]  69   8      1          1        0   164
#> [378,]  68   7      0        180        1   130
#> [379,]  72  16      1          1        1   130
#> [380,]  70   4      0        180        0   180
#> [381,]  73   6      1          0        1   270
#> [382,]  71   2      1          0        1   180
#> [383,]  73   7      0          7        1   140
#> [384,]  68  15      1         13        1   130
#> [385,]  70   3      0          3        1   159
#> [386,]  73   0      0        180        1   161
#> [387,]  74   8      1          0        1    85
#> [388,]  73   4      0        180        1   154
#> [389,]  71   3      1          1        0   150
#> [390,]  74  20      0         20        1   180
#> [391,]  74   0      1          0        1    90
#> [392,]  73   3      1          0        1   136
#> [393,]  70   5      1          0        1   190
#> [394,]  71  17      1         11        0   160
#> [395,]  71   8      1          7        0   149
#> [396,]  73  10      1          8        0   106
#> [397,]  69  12      1          1        1   149
#> [398,]  70  26      1         11        1   120
#> [399,]  74   4      0          4        0   120
#> [400,]  73   4      0         58        1   160
#> [401,]  72   5      1          3        1   160
#> [402,]  70   3      0        180        1   154
#> [403,]  71   7      1          2        0   143
#> [404,]  74   3      0          3        1   150
#> [405,]  73  17      1         11        0   140
#> [406,]  69   2      1          1        1    80
#> [407,]  70   4      1          0        1   140
#> [408,]  74   7      1          0        1   117
#> [409,]  69   7      0        180        1   144
#> [410,]  72  15      1         13        0   156
#> [411,]  71  10      1          9        1   120
#> [412,]  75   1      0          1        0   133
#> [413,]  73  10      1          9        1   146
#> [414,]  72  10      1          9        1   160
#> [415,]  71   2      0         10        1   112
#> [416,]  75   9      1          7        0   140
#> [417,]  75  13      1          1        1   130
#> [418,]  71  11      1          8        0   110
#> [419,]  71   4      0          4        0   134
#> [420,]  72  15      1         12        1   120
#> [421,]  73  10      1          8        0   120
#> [422,]  70   7      1          4        0   184
#> [423,]  72   1      1          1        0   168
#> [424,]  73  10      0        180        0   162
#> [425,]  72  11      0         11        1   140
#> [426,]  73   5      1          3        1   112
#> [427,]  72   2      0        180        0   120
#> [428,]  75   1      0        180        1   140
#> [429,]  72   4      1          0        1   197
#> [430,]  71   3      1          0        0   144
#> [431,]  73   5      0        180        0   126
#> [432,]  73   4      0        180        0   124
#> [433,]  76   3      1          0        1   120
#> [434,]  71  32      1         12        1   107
#> [435,]  72   5      0        180        0   154
#> [436,]  77  11      0         11        1   150
#> [437,]  77   4      0          4        0   185
#> [438,]  75   3      1          1        0   180
#> [439,]  73  15      0         15        1   160
#> [440,]  71  16      0        180        0   140
#> [441,]  73  10      1         10        0   124
#> [442,]  74   7      0        180        1   150
#> [443,]  74   3      0          3        1   128
#> [444,]  74   2      1          1        0   140
#> [445,]  76   8      1          0        1   141
#> [446,]  73   6      0          6        1   114
#> [447,]  75  23      1         14        1   110
#> [448,]  72   4      1          3        0   160
#> [449,]  76  17      1          0        1   200
#> [450,]  73   4      1          3        1   125
#> [451,]  76  13      1         10        0   110
#> [452,]  75   4      1          0        1   122
#> [453,]  75   0      0          0        1   130
#> [454,]  73  13      1         11        0   195
#> [455,]  75  12      0         12        1   160
#> [456,]  75   4      1          2        1   188
#> [457,]  76   4      0          4        1   155
#> [458,]  75   1      0          1        1   125
#> [459,]  74   2      0        180        0   111
#> [460,]  73   0      0        180        0   156
#> [461,]  76   5      0        180        0   185
#> [462,]  74  10      1          0        1   135
#> [463,]  76   5      1          0        1   167
#> [464,]  75   9      0        180        1   140
#> [465,]  77   5      1          0        0   123
#> [466,]  77  12      1          9        1   100
#> [467,]  73  10      1          9        0   146
#> [468,]  77  12      0        180        0   130
#> [469,]  77   1      1          0        1    90
#> [470,]  76  12      1         11        1   120
#> [471,]  73   7      1          0        0   174
#> [472,]  74   6      0         79        1   140
#> [473,]  75   3      1          1        1   171
#> [474,]  75   6      0        180        0   150
#> [475,]  74   2      1          0        1   130
#> [476,]  78  18      0         18        1   144
#> [477,]  76  29      0         47        0    90
#> [478,]  73  11      1          2        1   110
#> [479,]  78   7      0          7        1   133
#> [480,]  74  15      0        180        1   172
#> [481,]  78   8      1          6        1   110
#> [482,]  76  13      1          1        1   170
#> [483,]  79   6      0        180        0   170
#> [484,]  80  10      1          6        1   147
#> [485,]  78   0      0        180        1   212
#> [486,]  78  13      1          5        0   130
#> [487,]  78  15      0        180        1   270
#> [488,]  75  13      1          6        0   150
#> [489,]  76   1      0          1        1    83
#> [490,]  79   4      0         80        0   145
#> [491,]  78  12      1          9        0   150
#> [492,]  75   4      1          0        0   212
#> [493,]  77   2      1          0        1   143
#> [494,]  78  10      0        180        1   130
#> [495,]  76  11      1          0        0   120
#> [496,]  75  11      1          4        0   162
#> [497,]  75   3      0          3        0     0
#> [498,]  76   7      0         29        1   150
#> [499,]  79   8      0         32        1   120
#> [500,]  80   9      0         23        1   128
#> [501,]  80   6      0          6        1   150
#> [502,]  76   3      1          0        1   140
#> [503,]  78  11      1          1        1   140
#> [504,]  79  11      0        180        0   160
#> [505,]  78  14      1          0        1   140
#> [506,]  81   1      0          1        0   130
#> [507,]  76   4      0          4        1   160
#> [508,]  79   4      0          4        1   125
#> [509,]  76  10      1          8        0   180
#> [510,]  76  12      1         10        1   127
#> [511,]  77   6      0          6        1   107
#> [512,]  75   2      1          1        1   204
#> [513,]  78  11      0        180        1   135
#> [514,]  76   1      0          1        1    90
#> [515,]  78   7      1          0        1   110
#> [516,]  79   3      0          3        0   120
#> [517,]  77   7      0        180        1   170
#> [518,]  77   6      0          6        1   144
#> [519,]  79   4      1          0        1   120
#> [520,]  82   5      0          8        1   120
#> [521,]  80  40      1          0        1   138
#> [522,]  78   4      0         59        1   112
#> [523,]  80  17      1         12        0   100
#> [524,]  79  10      0         10        1   120
#> [525,]  79  28      0        164        0   100
#> [526,]  80   9      0        118        1   186
#> [527,]  78  32      0        180        1   130
#> [528,]  79   1      0         37        1   140
#> [529,]  81   3      0        180        0   184
#> [530,]  78   7      0          7        1   147
#> [531,]  77  13      1          0        1   190
#> [532,]  78  15      0         15        0   165
#> [533,]  80   5      1          1        1   108
#> [534,]  78   4      0        180        0   175
#> [535,]  79   3      0          3        1   101
#> [536,]  78  26      1          5        0   194
#> [537,]  81   4      1          1        1   104
#> [538,]  78  20      1          0        1   109
#> [539,]  82   3      1          1        1   144
#> [540,]  80   2      1          1        0   168
#> [541,]  79   6      0          6        0   152
#> [542,]  80   6      1          0        1   119
#> [543,]  78   2      0        180        0   148
#> [544,]  80   5      0          5        1   130
#> [545,]  79  10      0        180        1   150
#> [546,]  77   4      0        180        1    98
#> [547,]  78  12      0        180        0   134
#> [548,]  79   1      0        125        0   193
#> [549,]  84  22      1         10        0   180
#> [550,]  80   6      0          6        1   110
#> [551,]  83   9      1          5        1   170
#> [552,]  82   5      0        180        0   110
#> [553,]  83   5      0        180        0   148
#> [554,]  79   7      1          6        0   130
#> [555,]  81   5      0        177        0    41
#> [556,]  78  23      1         10        1   145
#> [557,]  79   4      0          4        1   183
#> [558,]  78   9      1          4        1   120
#> [559,]  82   8      1          1        0   128
#> [560,]  81  15      0        180        1   140
#> [561,]  81  20      1          9        0   170
#> [562,]  83   8      0          8        0   115
#> [563,]  80   6      1          0        1   150
#> [564,]  81   8      0        180        0   146
#> [565,]  79   7      0        177        0   197
#> [566,]  81   2      1          1        0   198
#> [567,]  83   2      0          2        1   155
#> [568,]  84   4      0        167        0   198
#> [569,]  81   1      0          1        1   150
#> [570,]  84   1      0         38        1   205
#> [571,]  83   3      0        180        0   174
#> [572,]  81   4      0         90        1   138
#> [573,]  79   9      1          8        0   150
#> [574,]  85   3      1          2        1   160
#> [575,]  80  13      1          8        1   140
#> [576,]  84   4      0         89        1   129
#> [577,]  80   2      1          0        1   130
#> [578,]  79   4      0          4        1    60
#> [579,]  80   6      0         71        1   189
#> [580,]  83   1      0          1        1   100
#> [581,]  82  19      0         19        0   120
#> [582,]  79  14      1          0        0   110
#> [583,]  83   3      0        114        0    98
#> [584,]  81  14      1         12        1   128
#> [585,]  83   2      0        154        0   130
#> [586,]  85   9      1          6        1   160
#> [587,]  83   1      0        180        0   160
#> [588,]  81   4      0          4        0   175
#> [589,]  84  15      1         13        1   110
#> [590,]  81  12      0         12        1   163
#> [591,]  82  16      1          8        0   103
#> [592,]  82   5      1          0        1   146
#> [593,]  81   4      0          4        0   160
#> [594,]  86  12      0        180        1   120
#> [595,]  81  19      1         14        0   120
#> [596,]  82   3      1          2        0   130
#> [597,]  82  15      1          0        0   183
#> [598,]  83   7      0        126        0   135
#> [599,]  86   8      0          8        1   132
#> [600,]  81  16      1          9        0   180
#> [601,]  84   6      0        165        0   145
#> [602,]  82   9      0        180        1   134
#> [603,]  84   3      0        180        1   120
#> [604,]  81  13      0        180        0   152
#> [605,]  85   3      0          3        1   118
#> [606,]  81   2      1          0        1   118
#> [607,]  81   4      0        180        0   160
#> [608,]  86  12      1          0        1   132
#> [609,]  86   6      1          0        1   140
#> [610,]  84  16      0         70        1   150
#> [611,]  84   3      1          2        0   125
#> [612,]  83  10      1          0        1   190
#> [613,]  86   2      0        180        1   169
#> [614,]  88  14      1          3        1   130
#> [615,]  84   3      0          3        1   121
#> [616,]  83  13      1         12        0   170
#> [617,]  84   7      1          2        0   148
#> [618,]  84   9      0         92        1   110
#> [619,]  84   3      0        180        1   170
#> [620,]  86  13      0        177        0   163
#> [621,]  84  13      0         62        1   100
#> [622,]  86   6      1          1        0   112
#> [623,]  88   4      0          4        0   100
#> [624,]  83  20      1          3        1   150
#> [625,]  83   9      0         65        1   150
#> [626,]  87   2      0        180        1   130
#> [627,]  83   3      0          3        1   130
#> [628,]  86  15      1          8        1   109
#> [629,]  89   4      0          4        1   153
#> [630,]  89   5      0        119        1   140
#> [631,]  84   8      0        180        1   119
#> [632,]  85   8      0          8        1   136
#> [633,]  84   2      0        110        1   174
#> [634,]  84   0      0        180        1   136
#> [635,]  90  14      0         14        1   100
#> [636,]  88   1      0          1        0   135
#> [637,]  91   8      0          8        0   100
#> [638,]  87   2      0        180        0   160
#> [639,]  86   3      1          0        1    80
#> [640,]  88   7      0         24        0   119
#> [641,]  88   8      0         50        1   154
#> [642,]  86  10      0        180        1   137
#> [643,]  86   9      1          7        0   130
#> [644,]  90   4      1          0        0   121
#> [645,]  87  43      0        178        1   130
#> [646,]  87   5      0         36        1   150
#> [647,]  92   1      0          1        1   167
#> [648,]  91   3      0         33        1   137
#> [649,]  87   7      0         74        1   105
#> [650,]  89  12      1          0        1   130
#> [651,]  91   5      0        169        1   176
#> [652,]  89  52      0         52        1   130
#> [653,]  92   7      0          7        1   110
#> [654,]  89   4      0          4        1   159
#> [655,]  91   0      0          0        0     0
#> [656,]  89  14      0        180        1    84
#> [657,]  91   4      1          0        1   120
#> [658,]  93   8      0        179        1   110
#> [659,]  94   8      0          8        1   142
#> [660,]  90   3      0         67        0   162
#> [661,]  96   3      0         12        1    97
#> [662,]  95   8      1          5        1   150
#> [663,]  94   3      0         26        1   144
#> [664,]  91  12      0         53        1   212
#> [665,]  91   7      0          7        0   135
#> [666,]  93   0      1          0        1   122
#> [667,]  92   5      0         69        0   139
#> [668,]  92   2      0          2        0   112
#> [669,]  93   4      0        180        1   135
#> [670,]  96  15      1          0        1   140
#> 
#> $y
#>   [1] 180.0+   5.0+   2.0+ 180.0+   5.0+ 180.0+   2.0+ 180.0+ 180.0+ 180.0+
#>  [11]   5.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0+ 180.0+   2.0+ 180.0+ 180.0+
#>  [21]   3.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 155.0+ 180.0+ 180.0+
#>  [31] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 150.0  180.0+ 180.0+   6.0+
#>  [41] 180.0+ 180.0+  73.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [51] 180.0+ 180.0+  10.0+ 172.0+ 180.0+ 180.0+ 180.0+   7.0  180.0+   2.0 
#>  [61]   1.0  179.0+ 180.0+ 179.0+ 180.0+ 180.0+   4.0+ 180.0+ 180.0+ 180.0+
#>  [71] 180.0+ 180.0+ 180.0+   7.0   36.0   88.0+ 180.0+   4.0+ 180.0+ 180.0+
#>  [81] 180.0+ 180.0+   5.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+
#>  [91]  99.0  180.0+  16.0+ 180.0+   7.0+   6.0+ 180.0+ 180.0+   6.0+   1.0 
#> [101] 180.0+   9.0+ 180.0+ 180.0+ 175.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [111] 180.0+ 180.0+ 180.0+  16.0   15.0+ 180.0+ 180.0+ 180.0+ 134.0+   8.0 
#> [121]   2.0    3.0+ 180.0+ 180.0+ 180.0+ 140.0  180.0+   1.0  180.0+ 180.0+
#> [131] 180.0+ 180.0+ 180.0+   8.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [141] 180.0+ 171.0+  15.0  166.0+   5.0+   4.0+ 180.0+   5.0+ 180.0+   4.0+
#> [151] 180.0+   9.0+ 180.0+   2.0+ 180.0+ 180.0+ 180.0+   2.0  180.0+ 180.0+
#> [161] 180.0+ 180.0+  64.0  180.0+ 180.0+ 180.0+   9.0+ 180.0+   0.5  161.0+
#> [171] 171.0+ 180.0+ 180.0+   1.0  180.0+ 180.0+ 180.0+  10.0+ 180.0+  17.0 
#> [181]  45.0    3.0+ 180.0+ 180.0+ 172.0+  24.0    8.0  180.0+ 180.0+   1.0+
#> [191] 180.0+ 180.0+  77.0   13.0+ 180.0+ 180.0+  94.0  180.0+ 169.0    7.0 
#> [201]   7.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+  13.0+   5.0   18.0    5.0+
#> [211] 180.0+ 180.0+   1.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+
#> [221] 180.0+ 180.0+   9.0  180.0+ 180.0+  84.0    7.0+ 180.0+ 180.0+ 180.0+
#> [231]   4.0+ 167.0  180.0+  17.0  180.0+  12.0  180.0+  36.0  180.0+  88.0 
#> [241]  12.0  180.0+ 180.0+   0.5  180.0+  12.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [251] 180.0+ 180.0+ 180.0+  18.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   3.0+
#> [261]   2.0+ 103.0   15.0  180.0+  13.0  180.0+ 179.0+ 166.0+   3.0    0.5+
#> [271] 180.0+ 180.0+ 180.0+   7.0+   8.0  180.0+ 180.0+ 180.0+ 180.0+ 123.0+
#> [281]   1.0+  18.0   11.0+ 180.0+  15.0+ 180.0+  15.0  180.0+ 180.0+ 174.0+
#> [291] 180.0+ 175.0  180.0+  10.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+  12.0 
#> [301]  99.0  179.0+ 180.0+ 180.0+  11.0+ 180.0+   0.5    7.0+ 180.0+ 180.0+
#> [311] 180.0+  93.0    1.0   18.0+   4.0    5.0    7.0+ 180.0+ 180.0+ 180.0+
#> [321]   1.0  180.0+ 171.0  180.0+ 174.0+   6.0  180.0+   0.5  180.0+ 180.0+
#> [331] 180.0+  14.0+   2.0  103.0    3.0+  36.0    5.0+ 180.0+   7.0    8.0+
#> [341] 180.0+ 180.0+  18.0  172.0+ 180.0+ 180.0+   7.0  180.0+   7.0   13.0+
#> [351] 180.0+  51.0   19.0    1.0   60.0   76.0  180.0+ 132.0  180.0+ 162.0 
#> [361]   7.0+ 124.0    9.0  180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 152.0    2.0 
#> [371] 173.0+ 180.0+ 180.0+  28.0  180.0+  16.0+ 180.0+ 180.0+  16.0+ 180.0+
#> [381]   6.0  180.0+   7.0+  15.0    3.0+ 180.0+ 180.0+ 180.0+   3.0+  20.0 
#> [391]   0.5  180.0+ 180.0+ 180.0+   8.0   87.0   12.0  180.0+   4.0+  58.0 
#> [401] 180.0+ 180.0+ 180.0+   3.0  180.0+   2.0  180.0+ 180.0+ 180.0+ 180.0+
#> [411] 179.0+   1.0  180.0+ 159.0   10.0  180.0+  13.0  180.0+   4.0+ 180.0+
#> [421]  10.0  104.0+   1.0  180.0+  11.0    5.0  180.0+ 180.0+ 180.0+ 180.0+
#> [431] 180.0+ 180.0+ 180.0+ 177.0+ 180.0+  11.0+   4.0+ 180.0+  15.0+ 180.0+
#> [441]  10.0  180.0+   3.0  180.0+ 180.0+   6.0  180.0+ 180.0+  17.0+ 180.0+
#> [451] 174.0+   4.0    0.5  180.0+  12.0   46.0    4.0    1.0  180.0+ 180.0+
#> [461] 180.0+ 180.0+ 180.0+ 180.0+   5.0  180.0+ 180.0+ 180.0+   1.0   12.0 
#> [471]   7.0+  79.0    3.0  180.0+ 176.0+  18.0   47.0   11.0    7.0  180.0+
#> [481]   8.0+ 180.0+ 180.0+  10.0  180.0+ 172.0  180.0+ 180.0+   1.0   80.0 
#> [491] 180.0+   4.0+   2.0  180.0+  11.0  152.0+   3.0   29.0   32.0   23.0 
#> [501]   6.0    3.0+ 180.0+ 180.0+ 180.0+   1.0    4.0    4.0   10.0+ 180.0+
#> [511]   6.0    2.0+ 180.0+   1.0   43.0    3.0  180.0+   6.0  138.0    8.0 
#> [521]  40.0   59.0   17.0   10.0+ 164.0  118.0  180.0+  37.0  180.0+   7.0+
#> [531]  22.0   15.0+   5.0+ 180.0+   3.0  171.0+  71.0   20.0+ 180.0+  10.0 
#> [541]   6.0+   6.0  180.0+   5.0  180.0+ 180.0+ 180.0+ 125.0  180.0+   6.0 
#> [551]   9.0+ 180.0+ 180.0+ 180.0+ 177.0+  70.0    4.0  180.0+ 180.0+ 180.0+
#> [561]  20.0    8.0+ 180.0+ 180.0+ 177.0+ 180.0+   2.0  167.0    1.0   38.0 
#> [571] 180.0+  90.0  180.0+ 180.0+ 180.0+  89.0  180.0+   4.0   71.0    1.0 
#> [581]  19.0  180.0+ 114.0  180.0+ 154.0  180.0+ 180.0+   4.0+ 180.0+  12.0 
#> [591]  16.0+   5.0+   4.0+ 180.0+ 180.0+   3.0   83.0  126.0    8.0  180.0+
#> [601] 165.0  180.0+ 180.0+ 180.0+   3.0+ 180.0+ 180.0+ 180.0+   6.0   70.0 
#> [611] 180.0+ 180.0+ 180.0+  14.0    3.0   13.0  180.0+  92.0  180.0+ 177.0 
#> [621]  62.0    6.0+   4.0+  20.0   65.0  180.0+   3.0+ 180.0+   4.0  119.0 
#> [631] 180.0+   8.0  110.0  180.0+  14.0    1.0+   8.0  180.0+   3.0   24.0 
#> [641]  50.0  180.0+ 180.0+   4.0  178.0+  36.0    1.0   33.0   74.0  180.0+
#> [651] 169.0   52.0    7.0    4.0    0.5  180.0+   4.0  179.0+   8.0+  67.0 
#> [661]  12.0    8.0   26.0   53.0    7.0+   0.5   69.0    2.0  180.0+  15.0+
#> 
#> $weights
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Warning: Multiple lambdas have been fit. Lambda will be set to 0.01 (see parameter 's').

# Score the predictions
predictions$score()
#> surv.cindex 
#>    0.832704