GLM with Elastic Net Regularization Survival Learner

Generalized linear models with elastic net regularization. Calls glmnet::glmnet() from package glmnet.

Initial parameter values

family is set to "cox" and cannot be changed.

Prediction types

This learner returns three prediction types:

lp: a vector containing the linear predictors (relative risk scores), where each score corresponds to a specific test observation. Calculated using glmnet::predict.coxnet().

crank: same as lp.

distr: a survival matrix in two dimensions, where observations are represented in rows and time points in columns. Calculated using glmnet::survfit.coxnet(). Parameters stype and ctype relate to how lp predictions are transformed into survival predictions and are described in survival::survfit.coxph(). By default the Breslow estimator is used for computing the baseline hazard.

Caution: This learner is different to learners calling glmnet::cv.glmnet() in that it does not use the internal optimization of parameter lambda. Instead, lambda needs to be tuned by the user (e.g., via mlr3tuning). When lambda is tuned, the glmnet will be trained for each tuning iteration. While fitting the whole path of lambdas would be more efficient, as is done by default in glmnet::glmnet(), tuning/selecting the parameter at prediction time (using parameter s) is currently not supported in mlr3 (at least not in efficient manner). Tuning the s parameter is, therefore, currently discouraged.

When the data are i.i.d. and efficiency is key, we recommend using the respective auto-tuning counterpart in mlr_learners_surv.cv_glmnet(). However, in some situations this is not applicable, usually when data are imbalanced or not i.i.d. (longitudinal, time-series) and tuning requires custom resampling strategies (blocked design, stratification).

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.coxnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

Dictionary

This Learner can be instantiated via lrn():

lrn("surv.glmnet")

Meta Information

Task type: “surv”

Predict Types: “crank”, “distr”, “lp”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3proba, mlr3extralearners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exact	logical	FALSE	TRUE, FALSE	-

exclude	untyped	-		-

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-Inf		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	TRUE	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	0.01		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.logistic	character	Newton	Newton, modified.Newton	-

type.multinomial	character	ungrouped	ungrouped, grouped	-

upper.limits	untyped	Inf		-

stype	integer	2		\([1, 2]\)

ctype	integer	-		\([1, 2]\)

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Author

be-marc

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvGlmnet

Methods

Public methods

LearnerSurvGlmnet$new()

LearnerSurvGlmnet$selected_features()

LearnerSurvGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerSurvGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner
learner = lrn("surv.glmnet")
print(learner)
#> 
#> ── <LearnerSurvGlmnet> (surv.glmnet): Regularized Generalized Linear Model ─────
#> • Model: -
#> • Parameters: use_pred_offset=TRUE
#> • Packages: mlr3, mlr3proba, mlr3extralearners, and glmnet
#> • Predict Types: [crank], distr, and lp
#> • Feature Types: logical, integer, and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: offset, selected_features, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("grace")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
#> $model
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "cox") 
#> 
#>    Df  %Dev   Lambda
#> 1   0  0.00 0.179500
#> 2   1  0.53 0.163500
#> 3   2  1.34 0.149000
#> 4   2  2.13 0.135800
#> 5   3  4.90 0.123700
#> 6   3  7.22 0.112700
#> 7   3  8.97 0.102700
#> 8   3 10.34 0.093580
#> 9   3 11.43 0.085260
#> 10  3 12.32 0.077690
#> 11  3 13.06 0.070790
#> 12  4 13.68 0.064500
#> 13  4 14.22 0.058770
#> 14  4 14.67 0.053550
#> 15  5 15.06 0.048790
#> 16  6 15.42 0.044460
#> 17  6 15.72 0.040510
#> 18  6 15.98 0.036910
#> 19  6 16.20 0.033630
#> 20  6 16.38 0.030640
#> 21  6 16.54 0.027920
#> 22  6 16.67 0.025440
#> 23  6 16.78 0.023180
#> 24  6 16.87 0.021120
#> 25  6 16.95 0.019240
#> 26  6 17.01 0.017530
#> 27  6 17.07 0.015980
#> 28  6 17.12 0.014560
#> 29  6 17.16 0.013260
#> 30  6 17.19 0.012090
#> 31  6 17.22 0.011010
#> 32  6 17.24 0.010030
#> 33  6 17.26 0.009142
#> 34  6 17.28 0.008330
#> 35  6 17.29 0.007590
#> 36  6 17.30 0.006916
#> 37  6 17.31 0.006301
#> 38  6 17.32 0.005742
#> 39  6 17.32 0.005232
#> 40  6 17.33 0.004767
#> 41  6 17.33 0.004343
#> 42  6 17.34 0.003958
#> 43  6 17.34 0.003606
#> 44  6 17.34 0.003286
#> 
#> $x
#>        age los revasc revascdays stchange sysbp
#>   [1,]  28   9      0        180        1   107
#>   [2,]  32   5      1          0        1   121
#>   [3,]  33   2      0          2        0   150
#>   [4,]  35   2      0        180        0   121
#>   [5,]  37   9      0        180        1   151
#>   [6,]  38   2      0        115        0   150
#>   [7,]  38  12      1          8        1   120
#>   [8,]  33   6      1          1        1   115
#>   [9,]  38  16      1         10        0   160
#>  [10,]  38  12      1         11        1    92
#>  [11,]  40  12      1          9        0   153
#>  [12,]  37   1      1          0        1   146
#>  [13,]  40   2      1          1        1   148
#>  [14,]  42   2      0        180        1   100
#>  [15,]  38   5      1          3        0   125
#>  [16,]  42   2      0          2        0   140
#>  [17,]  40   6      0        180        1   138
#>  [18,]  40  11      1         10        1   120
#>  [19,]  43   3      1          0        1   100
#>  [20,]  41   2      1          1        0   166
#>  [21,]  40   1      1          0        1   145
#>  [22,]  42   4      0        180        0   162
#>  [23,]  40   3      1          1        0   170
#>  [24,]  42  12      1         10        1   170
#>  [25,]  42   2      0        180        1   124
#>  [26,]  41  10      1          8        0   150
#>  [27,]  44   3      0        180        0   141
#>  [28,]  41  13      1          1        0   140
#>  [29,]  45   9      1          7        0   110
#>  [30,]  41   5      1          4        1   141
#>  [31,]  44   2      1          1        1   150
#>  [32,]  46  15      0        180        0   120
#>  [33,]  46   2      1          1        0   126
#>  [34,]  47   4      1          3        0   118
#>  [35,]  44   3      1          0        1   180
#>  [36,]  46   7      1          2        0   166
#>  [37,]  43  10      0        180        0   185
#>  [38,]  47   6      1          0        1   116
#>  [39,]  46  13      1         10        0   100
#>  [40,]  44   0      1          0        1    96
#>  [41,]  47   4      1          3        1   160
#>  [42,]  43   3      1          0        1   124
#>  [43,]  49   5      0         73        1   136
#>  [44,]  45   5      0          5        0   141
#>  [45,]  46   2      1          1        1   122
#>  [46,]  47   2      0        180        0   108
#>  [47,]  44   9      1          8        1   135
#>  [48,]  45   5      0        180        1   190
#>  [49,]  46   4      0        180        1   121
#>  [50,]  47   3      1          1        1   120
#>  [51,]  48   3      0        180        0   154
#>  [52,]  47   9      1          6        0   170
#>  [53,]  49   4      0        180        0   117
#>  [54,]  47  10      0         10        1   140
#>  [55,]  48   2      1          0        0   184
#>  [56,]  47   7      0        180        0   145
#>  [57,]  50   4      1          1        0   125
#>  [58,]  46   3      1          1        1   140
#>  [59,]  46   9      1          9        1   122
#>  [60,]  49   2      0          2        0   105
#>  [61,]  51   1      0          1        1   145
#>  [62,]  49  15      1         11        1   160
#>  [63,]  49  23      0        179        1   112
#>  [64,]  52   2      0        180        1   170
#>  [65,]  50   4      0          4        1   100
#>  [66,]  51   3      1          2        0   113
#>  [67,]  50   1      1          0        0   150
#>  [68,]  49   7      1          4        1    90
#>  [69,]  47   8      0        180        0   160
#>  [70,]  48   7      1          0        1   110
#>  [71,]  53   8      0         36        1   160
#>  [72,]  52   4      1          4        0   152
#>  [73,]  53   5      0        180        1   140
#>  [74,]  54  17      1         12        1   102
#>  [75,]  53   5      0         77        0   159
#>  [76,]  54   6      1          3        0   129
#>  [77,]  51   3      1          1        0   140
#>  [78,]  50   2      0          5        1   106
#>  [79,]  50  14      1         13        0   170
#>  [80,]  53   8      1          7        0   160
#>  [81,]  48   3      1          2        0   150
#>  [82,]  51  25      1          1        0   202
#>  [83,]  52  14      1          7        1   200
#>  [84,]  48   6      0        180        0   160
#>  [85,]  48  11      1         10        0   120
#>  [86,]  53   4      1          0        1   156
#>  [87,]  51  13      0         99        1   160
#>  [88,]  49  16      0         16        0   125
#>  [89,]  54  23      1         10        0   131
#>  [90,]  52   7      1          2        0   154
#>  [91,]  55   6      1          2        1   114
#>  [92,]  54   9      1          1        0   130
#>  [93,]  55   4      1          2        0   150
#>  [94,]  52   4      0        180        1   180
#>  [95,]  51  13      1         11        0   145
#>  [96,]  50   5      1          4        1   150
#>  [97,]  54   4      1          0        1   121
#>  [98,]  52   4      0        180        0   183
#>  [99,]  49   1      0          1        1   110
#> [100,]  50   7      1          1        0   156
#> [101,]  53   9      0          9        1    95
#> [102,]  53   8      1          0        1   130
#> [103,]  50   7      1          0        1   127
#> [104,]  55   1      0        180        0   127
#> [105,]  55   2      0          2        0   145
#> [106,]  54   7      1          0        1   100
#> [107,]  56   2      0        180        0   132
#> [108,]  55   5      1          4        1   120
#> [109,]  52  16      0         16        0   152
#> [110,]  53  10      1          9        0   172
#> [111,]  53  15      0         15        1    90
#> [112,]  53   4      0        180        1   150
#> [113,]  54  12      1          0        1   190
#> [114,]  55   2      0        134        1   140
#> [115,]  57   3      0          3        0   120
#> [116,]  54   7      1          2        0   129
#> [117,]  54   2      1          1        0   135
#> [118,]  52   9      1          3        0   170
#> [119,]  54   2      1          1        1   176
#> [120,]  57   5      1          3        1   138
#> [121,]  57   1      0        180        1   156
#> [122,]  57   1      0          1        1   100
#> [123,]  55  11      1          7        0   104
#> [124,]  52  15      1         14        0   130
#> [125,]  53   3      1          0        1   200
#> [126,]  57  10      0        180        1   170
#> [127,]  58   8      0          8        1   130
#> [128,]  54   5      0        180        1   108
#> [129,]  55   3      1          1        1   156
#> [130,]  53  21      1         13        1   130
#> [131,]  54  17      1          8        1   227
#> [132,]  55   9      1          2        1   147
#> [133,]  56   5      0          5        1   150
#> [134,]  57   4      1          2        1   185
#> [135,]  53   4      0        147        1   145
#> [136,]  57  11      1         10        1   129
#> [137,]  55   3      1          2        0   140
#> [138,]  54   7      1          0        1   141
#> [139,]  59  15      1         10        0   140
#> [140,]  58   9      1          0        1   180
#> [141,]  58   1      1          1        1   200
#> [142,]  55   5      1          0        0   140
#> [143,]  56   7      1          5        1   120
#> [144,]  55   2      0          2        0   106
#> [145,]  59   9      1          1        1   125
#> [146,]  57   1      0        180        0   148
#> [147,]  60  11      1          9        0   106
#> [148,]  57   2      0          2        1   120
#> [149,]  57   5      0        180        1   130
#> [150,]  55   5      1          0        1   160
#> [151,]  57  10      1          9        0   103
#> [152,]  59   6      1          0        1   140
#> [153,]  59   5      0        180        1   155
#> [154,]  59   4      1          0        1   152
#> [155,]  58  26      1          0        1   189
#> [156,]  61   9      0          9        1   160
#> [157,]  59   2      1          1        0   140
#> [158,]  58  14      1          6        0   190
#> [159,]  61   4      1          3        0   151
#> [160,]  61   9      1          8        0   150
#> [161,]  61   3      1          2        1   102
#> [162,]  58   1      0          1        1   100
#> [163,]  61  20      1         13        0   130
#> [164,]  57  13      1         10        0   110
#> [165,]  57   2      1          0        1   116
#> [166,]  58  10      0         10        1   150
#> [167,]  57   4      1          3        0   138
#> [168,]  61   3      0         17        0   143
#> [169,]  56  14      0         45        0   130
#> [170,]  57   3      1          2        0   120
#> [171,]  58  19      1         13        1   140
#> [172,]  56  13      1          6        1   158
#> [173,]  59   9      1          0        1    80
#> [174,]  55   4      1          3        1   160
#> [175,]  61   4      1          0        1   115
#> [176,]  56   8      1          8        0   120
#> [177,]  59  11      1          8        1   190
#> [178,]  57   1      0          1        0   126
#> [179,]  57  15      1         13        1   110
#> [180,]  61   8      0         77        0   120
#> [181,]  58   8      1          5        0   152
#> [182,]  62  10      1          0        1   153
#> [183,]  57   3      1          0        0   100
#> [184,]  61  18      0        170        0   140
#> [185,]  61  28      1          7        0   133
#> [186,]  58   8      1          3        1   150
#> [187,]  60   7      0          7        0   147
#> [188,]  61   6      0          6        0   134
#> [189,]  59  13      1          2        0   198
#> [190,]  57  12      1          9        1   120
#> [191,]  62   4      1          0        0   160
#> [192,]  58   3      1          0        1   146
#> [193,]  63   6      0         28        1   120
#> [194,]  61   5      0          5        1   110
#> [195,]  57  18      1          9        1    93
#> [196,]  57   2      1          1        0   159
#> [197,]  62  17      1         10        1   180
#> [198,]  62   1      1          0        1   172
#> [199,]  58   7      0        180        1   150
#> [200,]  63   3      1          1        0   180
#> [201,]  63   1      0        180        1   130
#> [202,]  61   7      0        180        0   135
#> [203,]  63   4      1          3        0   222
#> [204,]  62   3      0        180        1   105
#> [205,]  63   4      0        180        1   190
#> [206,]  63  15      1         10        1   126
#> [207,]  64   4      0        180        0   130
#> [208,]  63   4      1          1        0   155
#> [209,]  60  18      1         13        0   132
#> [210,]  59   8      0        180        1   140
#> [211,]  58   9      1          9        0   110
#> [212,]  59   1      0         22        1   162
#> [213,]  58   2      0        180        0   127
#> [214,]  59   4      0        180        0   196
#> [215,]  60   7      1          5        1   141
#> [216,]  60   7      1          1        1    90
#> [217,]  63   1      0          1        0   162
#> [218,]  63   1      0          1        0   130
#> [219,]  62   6      0        180        0   170
#> [220,]  61  15      1         13        0   170
#> [221,]  64  10      1          9        0   160
#> [222,]  63  12      1         10        0   200
#> [223,]  59  10      0        180        1   130
#> [224,]  60   8      0         17        1   130
#> [225,]  61   6      1          1        1   117
#> [226,]  64  12      1         11        0   160
#> [227,]  65  36      1         11        0   140
#> [228,]  66   3      1          1        0   127
#> [229,]  61  10      1          2        1   194
#> [230,]  64  32      1          9        1   160
#> [231,]  63  12      1          9        0   114
#> [232,]  63   7      0        180        0   120
#> [233,]  66   5      1          0        1   110
#> [234,]  65   8      1          0        0   168
#> [235,]  65  10      1          8        1   120
#> [236,]  64   0      0          0        1    90
#> [237,]  60   6      0        180        0   130
#> [238,]  64  21      1         10        0   190
#> [239,]  61  12      1         11        0   154
#> [240,]  64   9      0        180        0   150
#> [241,]  65   3      0        180        1   190
#> [242,]  63  16      1          7        1   110
#> [243,]  64   7      0        180        1   120
#> [244,]  66   6      1          1        1   130
#> [245,]  65   6      0          9        0   112
#> [246,]  65   3      1          0        1    80
#> [247,]  63   5      1          4        0   170
#> [248,]  62  13      1         11        0   180
#> [249,]  66  18      1          5        0   142
#> [250,]  66  16      1         11        1   169
#> [251,]  62   9      0        180        0   145
#> [252,]  61  14      1          5        0   140
#> [253,]  61  15      1         10        0   130
#> [254,]  63   3      1          2        0   120
#> [255,]  64  19      1          8        1   160
#> [256,]  65   8      1          0        1   140
#> [257,]  67   6      0        180        1   170
#> [258,]  65  15      1         11        1   160
#> [259,]  64   6      1          0        1   125
#> [260,]  66  13      1          0        0   118
#> [261,]  64  14      1         13        1   150
#> [262,]  64   0      0          0        1   148
#> [263,]  67   4      1          3        0   130
#> [264,]  66   3      1          0        1   135
#> [265,]  66   6      1          0        1   140
#> [266,]  68   1      0        180        1   166
#> [267,]  64  10      1          9        1   110
#> [268,]  63   7      1          0        0   162
#> [269,]  67   8      1          1        1   130
#> [270,]  66  14      0        180        0   130
#> [271,]  65  17      1         14        1   100
#> [272,]  63   8      1          1        1   162
#> [273,]  65  18      1          3        0   120
#> [274,]  63   1      1          0        1   155
#> [275,]  67  11      0         11        0   150
#> [276,]  68  14      0         79        0   172
#> [277,]  65  15      1         12        1   150
#> [278,]  66  15      1         13        1   160
#> [279,]  63   2      0        180        0   150
#> [280,]  65  11      1          6        0   130
#> [281,]  69  21      1         10        0   180
#> [282,]  69   6      0        180        1   100
#> [283,]  66   9      1          8        0   130
#> [284,]  68  14      1         13        1   140
#> [285,]  65   8      1          0        1    90
#> [286,]  66   3      0          3        1   138
#> [287,]  69   1      1          0        0   170
#> [288,]  67   1      0        180        1   160
#> [289,]  68  10      1         10        1   150
#> [290,]  65   1      1          0        0   133
#> [291,]  67   7      1          4        1   130
#> [292,]  63   2      1          0        0    99
#> [293,]  65   6      0          6        0    80
#> [294,]  65  10      1          1        1   148
#> [295,]  66  19      1         12        1   150
#> [296,]  67  12      1         12        0   160
#> [297,]  69   6      0         99        1   140
#> [298,]  65   4      1          1        0   130
#> [299,]  64   4      0        179        0   160
#> [300,]  70  15      1         12        1   132
#> [301,]  64  11      0         11        0   125
#> [302,]  64   4      0        180        1   140
#> [303,]  64   0      1          0        1   118
#> [304,]  67   2      0         18        0   131
#> [305,]  66   4      0        180        0   177
#> [306,]  68   4      1          0        1   160
#> [307,]  65  13      1         12        1   130
#> [308,]  69   8      0         93        0   140
#> [309,]  66   6      0        180        0   140
#> [310,]  65   6      0        101        1   115
#> [311,]  68   4      0          4        1   190
#> [312,]  71   3      0          5        0   112
#> [313,]  70   7      1          0        1   190
#> [314,]  71  20      1          0        1   160
#> [315,]  67   2      0        180        0   128
#> [316,]  66   9      1          3        1   151
#> [317,]  66   1      1          1        1   165
#> [318,]  70   4      1          0        1   180
#> [319,]  70  14      0        171        0   166
#> [320,]  67  10      1          9        0   200
#> [321,]  67   6      1          4        0   130
#> [322,]  68  18      1         14        1   170
#> [323,]  69   0      0          0        1   148
#> [324,]  65   2      0        180        0   130
#> [325,]  68   7      1          0        1   150
#> [326,]  65  14      1         13        1   150
#> [327,]  69   8      0        180        1   180
#> [328,]  71   7      0          7        0   230
#> [329,]  71   6      0         45        1   158
#> [330,]  69   3      0          3        1   130
#> [331,]  70  22      1         13        0   103
#> [332,]  67   1      0         36        1   104
#> [333,]  67   5      0          5        0   130
#> [334,]  69   8      1          5        1   195
#> [335,]  69   6      1          4        1   174
#> [336,]  72   3      1          0        1   132
#> [337,]  72   7      0          7        1   110
#> [338,]  69   8      1          7        1   108
#> [339,]  67   3      0        180        0   110
#> [340,]  66   2      1          1        0   123
#> [341,]  66   2      0        180        0   130
#> [342,]  67   7      1          4        0   122
#> [343,]  67  13      1          9        0   130
#> [344,]  72   5      1          4        0   170
#> [345,]  67  12      1          8        0   120
#> [346,]  67   1      0          1        1    60
#> [347,]  67   4      0         60        1   136
#> [348,]  69   5      0         76        0   120
#> [349,]  67   8      1          0        1   130
#> [350,]  72  13      1         11        1   195
#> [351,]  68  10      1          8        1   160
#> [352,]  66  24      1         13        0   130
#> [353,]  72  30      1          0        1   145
#> [354,]  70   7      0          7        0   102
#> [355,]  68   7      1          2        0   135
#> [356,]  73  20      1          0        1   170
#> [357,]  71   6      0          9        0   120
#> [358,]  69  10      1          6        1   120
#> [359,]  70  11      0        180        1   210
#> [360,]  72  19      1          8        0   120
#> [361,]  72  12      1         10        0   170
#> [362,]  67   8      0        180        1   170
#> [363,]  67   5      1          0        1   147
#> [364,]  67   9      0        180        0   158
#> [365,]  73  13      0        152        1   130
#> [366,]  70   5      0        180        0   150
#> [367,]  72   6      1          5        0   115
#> [368,]  71   1      0        173        1   188
#> [369,]  68  23      0        180        1   220
#> [370,]  70   3      0        180        0   121
#> [371,]  69   3      0        180        0   220
#> [372,]  71   3      1          2        0   150
#> [373,]  68   4      1          3        0   210
#> [374,]  72   5      0         28        0   120
#> [375,]  73   6      0        180        1   117
#> [376,]  69  16      1         10        1   140
#> [377,]  69   8      1          1        0   164
#> [378,]  68   7      0        180        1   130
#> [379,]  72  16      1          1        1   130
#> [380,]  70   4      0        180        0   180
#> [381,]  69   1      1          0        0   155
#> [382,]  73   6      1          0        1   270
#> [383,]  72   8      1          1        1   150
#> [384,]  73   7      0          7        1   140
#> [385,]  70   3      0          3        1   159
#> [386,]  70  13      1          9        0   100
#> [387,]  72   6      0        180        1   130
#> [388,]  73   0      0        180        1   161
#> [389,]  74   8      1          0        1    85
#> [390,]  71  15      1         11        0   165
#> [391,]  71  20      1         10        0   140
#> [392,]  74   0      1          0        1    90
#> [393,]  73   3      1          0        1   136
#> [394,]  71  17      1         11        0   160
#> [395,]  71   3      1          2        1   190
#> [396,]  73  10      1          8        0   106
#> [397,]  69  12      1          1        1   149
#> [398,]  70  26      1         11        1   120
#> [399,]  73   4      0         58        1   160
#> [400,]  70   3      0        180        1   154
#> [401,]  73   6      0        180        0   110
#> [402,]  72   8      1          0        1   140
#> [403,]  74   3      0          3        1   150
#> [404,]  69   2      1          1        1    80
#> [405,]  70   4      1          0        1   140
#> [406,]  74   7      1          0        1   117
#> [407,]  70   8      0          8        0   120
#> [408,]  75   1      0          1        0   133
#> [409,]  75   2      1          1        0   145
#> [410,]  73  10      1          9        1   146
#> [411,]  72  10      1          9        1   160
#> [412,]  73  10      1         10        1   120
#> [413,]  71   2      0         10        1   112
#> [414,]  75   9      1          7        0   140
#> [415,]  75  13      1          1        1   130
#> [416,]  72  15      1         12        1   120
#> [417,]  73  10      1          8        0   120
#> [418,]  70   7      1          4        0   184
#> [419,]  72   1      1          1        0   168
#> [420,]  72   7      0         57        1   145
#> [421,]  73  10      0        180        0   162
#> [422,]  70   3      0          3        0   150
#> [423,]  73   5      1          3        1   112
#> [424,]  73  12      1         12        1   140
#> [425,]  72   4      1          0        1   197
#> [426,]  73   5      0        180        0   126
#> [427,]  74  34      1          8        1   233
#> [428,]  76   3      1          0        1   120
#> [429,]  71  32      1         12        1   107
#> [430,]  72   5      0        180        0   154
#> [431,]  76   5      0          5        1   130
#> [432,]  77   4      0          4        0   185
#> [433,]  72   7      1          2        0   142
#> [434,]  73  15      0         15        1   160
#> [435,]  71  16      0        180        0   140
#> [436,]  73  10      1         10        0   124
#> [437,]  74   3      0          3        1   128
#> [438,]  76   1      0        180        0   114
#> [439,]  76   8      1          0        1   141
#> [440,]  74  19      1          4        1   200
#> [441,]  73   6      0          6        1   114
#> [442,]  75  23      1         14        1   110
#> [443,]  74   2      0        180        0   190
#> [444,]  72   4      0         85        1   120
#> [445,]  72   4      1          3        0   160
#> [446,]  76  17      1          0        1   200
#> [447,]  73   4      1          3        1   125
#> [448,]  75   4      1          0        1   122
#> [449,]  73  13      1         11        0   195
#> [450,]  74   6      0        180        0   160
#> [451,]  76   4      0          4        1   155
#> [452,]  74   2      0        180        0   111
#> [453,]  73   1      0         52        1   105
#> [454,]  72   5      0        180        0   120
#> [455,]  76  44      1         10        0   105
#> [456,]  74  10      1          0        1   135
#> [457,]  76   5      1          0        1   167
#> [458,]  74   8      1          8        1   170
#> [459,]  75   9      0        180        1   140
#> [460,]  77   5      1          0        0   123
#> [461,]  77  12      1          9        1   100
#> [462,]  73  10      1          9        0   146
#> [463,]  77  12      0        180        0   130
#> [464,]  77   1      1          0        1    90
#> [465,]  78   5      1          0        1   170
#> [466,]  73   7      1          0        0   174
#> [467,]  74   6      0         79        1   140
#> [468,]  75   6      0        180        0   150
#> [469,]  79  10      1          8        0   190
#> [470,]  78  18      0         18        1   144
#> [471,]  77   3      0        180        0   110
#> [472,]  73   8      1          1        1   162
#> [473,]  73  11      1          2        1   110
#> [474,]  74   2      0        180        0   100
#> [475,]  78   7      0          7        1   133
#> [476,]  74  15      0        180        1   172
#> [477,]  78   8      1          6        1   110
#> [478,]  74   7      0          7        0   161
#> [479,]  76  13      1          1        1   170
#> [480,]  80  10      1          6        1   147
#> [481,]  78  13      1          5        0   130
#> [482,]  75   5      0        119        1   150
#> [483,]  75  12      1          1        1   120
#> [484,]  80   8      0          8        1   120
#> [485,]  75  13      1          6        0   150
#> [486,]  79   4      0         80        0   145
#> [487,]  78  12      1          9        0   150
#> [488,]  78   2      1          1        0   130
#> [489,]  75   4      1          0        0   212
#> [490,]  77   2      1          0        1   143
#> [491,]  78  10      0        180        1   130
#> [492,]  76  11      1          0        0   120
#> [493,]  75  11      1          4        0   162
#> [494,]  76   7      0         29        1   150
#> [495,]  77  24      0         24        1   160
#> [496,]  80   6      0          6        1   150
#> [497,]  76   3      1          0        1   140
#> [498,]  78  11      1          1        1   140
#> [499,]  79  11      0        180        0   160
#> [500,]  78  14      1          0        1   140
#> [501,]  81   1      0          1        0   130
#> [502,]  76   4      0          4        1   160
#> [503,]  79   4      0          4        1   125
#> [504,]  76  12      1         10        1   127
#> [505,]  75   2      1          1        1   204
#> [506,]  76   1      0          1        1   140
#> [507,]  77  31      1          3        1   161
#> [508,]  79   3      0          3        0   120
#> [509,]  77   6      0          6        1   144
#> [510,]  79   4      1          0        1   120
#> [511,]  81   1      0        180        0   120
#> [512,]  77   9      1          4        0   141
#> [513,]  82   5      0          8        1   120
#> [514,]  80  17      1         12        0   100
#> [515,]  76   7      0        161        0   151
#> [516,]  79  10      0         10        1   120
#> [517,]  79  28      0        164        0   100
#> [518,]  80   6      0        173        1   160
#> [519,]  78  32      0        180        1   130
#> [520,]  79   1      0         37        1   140
#> [521,]  81   3      0        180        0   184
#> [522,]  78   7      0          7        1   147
#> [523,]  77  13      1          0        1   190
#> [524,]  78  15      0         15        0   165
#> [525,]  78   4      0        180        0   175
#> [526,]  78  26      1          5        0   194
#> [527,]  76   1      0        166        0   131
#> [528,]  80   1      0          1        0   100
#> [529,]  78   3      1          1        1   152
#> [530,]  82   3      1          1        1   144
#> [531,]  80   2      1          1        0   168
#> [532,]  80   6      1          0        1   119
#> [533,]  78   2      0        180        0   148
#> [534,]  80   5      0          5        1   130
#> [535,]  82   1      1          0        1    82
#> [536,]  82  21      1          2        0   155
#> [537,]  84  22      1         10        0   180
#> [538,]  79   4      0          4        1   121
#> [539,]  80   6      0          6        1   110
#> [540,]  83   9      1          5        1   170
#> [541,]  82   5      0        180        0   110
#> [542,]  83   5      0        180        0   148
#> [543,]  83   4      0        103        0    97
#> [544,]  81   5      0        177        0    41
#> [545,]  78  23      1         10        1   145
#> [546,]  79   4      0          4        1   183
#> [547,]  79   1      0        180        1   170
#> [548,]  80   7      1          0        1   146
#> [549,]  84   5      1          1        1    85
#> [550,]  83   8      0          8        0   115
#> [551,]  81  16      0         16        1   110
#> [552,]  80   6      1          0        1   150
#> [553,]  80  11      1          8        0   110
#> [554,]  81   8      0        180        0   146
#> [555,]  80   8      1          7        0   160
#> [556,]  79   7      0        177        0   197
#> [557,]  85   4      0        180        0    90
#> [558,]  83   2      0          2        1   155
#> [559,]  84   4      0        167        0   198
#> [560,]  80   3      1          1        1   230
#> [561,]  82  23      1          0        0   110
#> [562,]  84   5      0        180        1   203
#> [563,]  84   4      0          4        1    85
#> [564,]  81   1      0          1        1   150
#> [565,]  84   1      0         38        1   205
#> [566,]  79   9      1          8        0   150
#> [567,]  85   3      1          2        1   160
#> [568,]  80  13      1          8        1   140
#> [569,]  84   4      0         89        1   129
#> [570,]  80   2      1          0        1   130
#> [571,]  79   4      0          4        1    60
#> [572,]  80   6      0         71        1   189
#> [573,]  80  30      1         13        0   220
#> [574,]  83   9      0        180        0   198
#> [575,]  79  14      1          0        0   110
#> [576,]  83   3      0        114        0    98
#> [577,]  83   2      0        154        0   130
#> [578,]  85   9      1          6        1   160
#> [579,]  81   4      0          4        0   175
#> [580,]  84  15      1         13        1   110
#> [581,]  81   1      0          1        1   145
#> [582,]  82  16      1          8        0   103
#> [583,]  81   4      0          4        0   160
#> [584,]  86  12      0        180        1   120
#> [585,]  86   8      0          8        1   132
#> [586,]  81  16      1          9        0   180
#> [587,]  84   6      0        165        0   145
#> [588,]  86   3      0          3        1   140
#> [589,]  82   9      0        180        1   134
#> [590,]  84   3      0        180        1   120
#> [591,]  81  13      0        180        0   152
#> [592,]  85   3      0          3        1   118
#> [593,]  81   2      1          0        1   118
#> [594,]  81   4      0        180        0   160
#> [595,]  83   9      0        180        1   149
#> [596,]  82   1      0        180        1   193
#> [597,]  83   4      0          4        0   130
#> [598,]  87   2      0          5        1   137
#> [599,]  82  14      1         11        1   103
#> [600,]  83  19      0         43        0   150
#> [601,]  84   3      1          2        0   125
#> [602,]  83  10      1          0        1   190
#> [603,]  86   2      0        180        1   169
#> [604,]  88  14      1          3        1   130
#> [605,]  84   3      0          3        1   121
#> [606,]  83  13      1         12        0   170
#> [607,]  84   9      0         92        1   110
#> [608,]  82   4      0          4        0   130
#> [609,]  86  13      0        177        0   163
#> [610,]  85   3      0          3        1   113
#> [611,]  86   6      0          6        1   117
#> [612,]  84  13      0         62        1   100
#> [613,]  86   6      1          1        0   112
#> [614,]  83  20      1          3        1   150
#> [615,]  88   4      0          4        1   115
#> [616,]  83   9      0         65        1   150
#> [617,]  87   2      0        180        1   130
#> [618,]  86   6      0         46        0   173
#> [619,]  88   3      0        115        0   110
#> [620,]  88   2      0        180        1    68
#> [621,]  83   3      0          3        1   130
#> [622,]  87   8      0          8        1   157
#> [623,]  86  15      1          8        1   109
#> [624,]  88   4      0          4        0    86
#> [625,]  89   4      0          4        1   153
#> [626,]  89   5      0        119        1   140
#> [627,]  87   6      0        180        1   110
#> [628,]  84   8      0        180        1   119
#> [629,]  85   8      0          8        1   136
#> [630,]  84   2      0        110        1   174
#> [631,]  89  10      0         46        1   170
#> [632,]  90  14      0         14        1   100
#> [633,]  88   1      0          1        0   135
#> [634,]  91  10      0        145        0   135
#> [635,]  86   3      1          0        1    80
#> [636,]  88   7      0         24        0   119
#> [637,]  88   8      0         50        1   154
#> [638,]  87   6      0        126        1   168
#> [639,]  86   9      1          7        0   130
#> [640,]  91   1      0          1        1    74
#> [641,]  87  43      0        178        1   130
#> [642,]  90   5      1          0        1   125
#> [643,]  88   3      1          2        0   159
#> [644,]  89   3      1          1        1   160
#> [645,]  92   1      0          1        1   167
#> [646,]  91   3      0         33        1   137
#> [647,]  89  12      1          0        1   130
#> [648,]  89   2      0        168        0   118
#> [649,]  89  52      0         52        1   130
#> [650,]  92   7      0          7        1   110
#> [651,]  89   4      0          4        1   159
#> [652,]  91   0      0          0        0     0
#> [653,]  89  14      0        180        1    84
#> [654,]  90  18      0        180        0   188
#> [655,]  91   4      1          0        1   120
#> [656,]  90  19      1         11        1   129
#> [657,]  94   6      0         50        0    78
#> [658,]  93   8      0        179        1   110
#> [659,]  91   1      0        180        0   158
#> [660,]  90  16      0         16        1   106
#> [661,]  90   3      0         67        0   162
#> [662,]  96   3      0         12        1    97
#> [663,]  95   8      1          5        1   150
#> [664,]  94   3      0         26        1   144
#> [665,]  91   7      0          7        0   135
#> [666,]  93   0      1          0        1   122
#> [667,]  92   2      0          2        0   112
#> [668,]  93   4      0        180        1   135
#> [669,]  96   3      1          0        1   104
#> [670,]  96  15      1          0        1   140
#> 
#> $y
#>   [1] 180.0+   5.0+   2.0+ 180.0+ 180.0+ 115.0   12.0  180.0+ 180.0+ 180.0+
#>  [11] 180.0+ 180.0+   2.0+ 180.0+   5.0+   2.0+ 180.0+ 180.0+   3.0  180.0+
#>  [21] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   5.0+
#>  [31] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   6.0+ 180.0+ 180.0+
#>  [41] 180.0+ 180.0+  73.0    5.0+ 161.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+
#>  [51] 180.0+ 180.0+ 180.0+  10.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   2.0 
#>  [61]   1.0  179.0+ 179.0+ 180.0+   4.0+ 180.0+ 180.0+ 180.0+ 180.0+   7.0 
#>  [71]  36.0    4.0+ 180.0+ 180.0+  77.0  180.0+ 180.0+   5.0  180.0+ 180.0+
#>  [81] 180.0+ 180.0+  85.0  180.0+ 180.0+ 166.0+  99.0   16.0+ 152.0+   7.0+
#>  [91]   6.0+ 180.0+ 180.0+ 180.0+  13.0+ 171.0+ 180.0+ 180.0+   1.0  180.0+
#> [101]   9.0+ 180.0+ 180.0+ 180.0+   2.0    7.0+ 180.0+ 180.0+  16.0+ 180.0+
#> [111]  15.0+ 180.0+  12.0+ 134.0+   3.0+ 180.0+ 180.0+ 180.0+ 180.0+ 140.0 
#> [121] 180.0+   1.0  180.0+ 180.0+ 180.0+ 180.0+   8.0+ 180.0+ 180.0+ 180.0+
#> [131] 171.0+  15.0    5.0+   4.0+ 147.0+ 180.0+ 180.0+ 180.0+ 180.0+   9.0+
#> [141]   1.0  180.0+ 180.0+   2.0+ 180.0+ 180.0+ 180.0+   2.0  180.0+ 180.0+
#> [151] 180.0+  64.0  180.0+ 180.0+ 180.0+   9.0+ 180.0+ 171.0+ 180.0+ 180.0+
#> [161]   3.0    1.0  180.0+ 180.0+ 180.0+  10.0+ 180.0+  17.0   45.0    3.0+
#> [171]  19.0  180.0+   9.0+ 180.0+ 180.0+   8.0  180.0+   1.0+  15.0   77.0 
#> [181]   8.0+ 180.0+ 180.0+ 170.0   94.0  180.0+   7.0+   6.0  180.0+ 180.0+
#> [191] 180.0+   3.0+  28.0    5.0   18.0  180.0+ 180.0+   1.0  180.0+ 180.0+
#> [201] 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0+ 180.0+ 180.0+
#> [211]   9.0   22.0  180.0+ 180.0+  84.0  180.0+   1.0    1.0  180.0+ 180.0+
#> [221] 167.0  180.0+ 180.0+  17.0  180.0+  12.0   36.0    3.0+  88.0  180.0+
#> [231]  12.0  180.0+ 180.0+ 180.0+ 180.0+   0.5  180.0+ 180.0+  12.0+ 180.0+
#> [241] 180.0+ 180.0+ 180.0+ 180.0+   9.0    3.0  180.0+ 180.0+  18.0+ 180.0+
#> [251] 180.0+ 180.0+ 180.0+   3.0+ 103.0   15.0  180.0+ 180.0+ 180.0+ 166.0+
#> [261]  14.0+   0.5+ 180.0+   3.0+ 180.0+ 180.0+ 180.0+   7.0+   8.0  180.0+
#> [271] 180.0+ 180.0+ 123.0+   1.0+  11.0+  79.0   15.0+ 180.0+ 180.0+ 180.0+
#> [281] 174.0+ 180.0+ 180.0+ 180.0+   8.0+   3.0  175.0  180.0+  10.0  180.0+
#> [291] 180.0+ 180.0+   6.0  180.0+  19.0+  12.0   99.0  180.0+ 179.0+ 180.0+
#> [301]  11.0+ 180.0+   0.5   18.0  180.0+ 180.0+ 180.0+  93.0  180.0+ 101.0 
#> [311]   4.0    5.0    7.0+ 180.0+ 180.0+ 180.0+   1.0  180.0+ 171.0  174.0+
#> [321]   6.0  180.0+   0.5  180.0+ 180.0+  14.0+ 180.0+   7.0+  45.0    3.0+
#> [331] 180.0+  36.0    5.0+ 180.0+  97.0  180.0+   7.0    8.0+ 180.0+   2.0+
#> [341] 180.0+   7.0   13.0+ 180.0+ 180.0+   1.0   60.0   76.0  180.0+ 132.0 
#> [351]  10.0+ 180.0+ 162.0    7.0+   7.0+ 124.0    9.0  180.0+ 180.0+ 180.0+
#> [361]  12.0  180.0+ 180.0+ 180.0+ 152.0  180.0+ 180.0+ 173.0+ 180.0+ 180.0+
#> [371] 180.0+ 180.0+ 180.0+  28.0  180.0+  16.0+ 180.0+ 180.0+  16.0+ 180.0+
#> [381] 180.0+   6.0  180.0+   7.0+   3.0+  13.0+ 180.0+ 180.0+ 180.0+ 180.0+
#> [391]  20.0    0.5  180.0+ 180.0+   3.0   87.0   12.0  180.0+  58.0  180.0+
#> [401] 180.0+ 180.0+   3.0    2.0  180.0+ 180.0+   8.0+   1.0  180.0+ 180.0+
#> [411] 159.0   15.0   10.0  180.0+  13.0  180.0+  10.0  104.0+   1.0   57.0 
#> [421] 180.0+   3.0+   5.0   12.0  180.0+ 180.0+  34.0  180.0+ 177.0+ 180.0+
#> [431]   5.0    4.0+   7.0   15.0+ 180.0+  10.0    3.0  180.0+ 180.0+ 180.0+
#> [441]   6.0  180.0+ 180.0+  85.0  180.0+  17.0+ 180.0+   4.0  180.0+ 180.0+
#> [451]   4.0  180.0+  52.0  180.0+ 180.0+ 180.0+ 180.0+   8.0  180.0+   5.0 
#> [461] 180.0+ 180.0+ 180.0+   1.0  180.0+   7.0+  79.0  180.0+ 180.0+  18.0 
#> [471] 180.0+ 180.0+  11.0  180.0+   7.0  180.0+   8.0+   7.0  180.0+  10.0 
#> [481] 172.0  119.0   12.0    8.0  180.0+  80.0  180.0+ 180.0+   4.0+   2.0 
#> [491] 180.0+  11.0  152.0+  29.0   24.0    6.0    3.0+ 180.0+ 180.0+ 180.0+
#> [501]   1.0    4.0    4.0  180.0+   2.0+   1.0  171.0    3.0    6.0  138.0 
#> [511] 180.0+  71.0    8.0   17.0  161.0   10.0+ 164.0  173.0  180.0+  37.0 
#> [521] 180.0+   7.0+  22.0   15.0+ 180.0+ 171.0+ 166.0+   1.0    3.0+ 180.0+
#> [531]  10.0    6.0  180.0+   5.0    1.0  180.0+ 180.0+   4.0    6.0    9.0+
#> [541] 180.0+ 180.0+ 103.0  177.0+  70.0    4.0  180.0+   7.0+ 180.0+   8.0+
#> [551]  16.0  180.0+ 180.0+ 180.0+ 180.0+ 177.0+ 180.0+   2.0  167.0    3.0+
#> [561]  62.0  180.0+   4.0    1.0   38.0  180.0+ 180.0+ 180.0+  89.0  180.0+
#> [571]   4.0   71.0   30.0  180.0+ 180.0+ 114.0  154.0  180.0+   4.0+ 180.0+
#> [581]   1.0   16.0+   4.0+ 180.0+   8.0  180.0+ 165.0    3.0  180.0+ 180.0+
#> [591] 180.0+   3.0+ 180.0+ 180.0+ 180.0+ 180.0+   4.0+   5.0  174.0   43.0 
#> [601] 180.0+ 180.0+ 180.0+  14.0    3.0   13.0   92.0    4.0  177.0    3.0+
#> [611]   6.0+  62.0    6.0+  20.0    4.0   65.0  180.0+  46.0  115.0  180.0+
#> [621]   3.0+   8.0+ 180.0+   4.0    4.0  119.0  180.0+ 180.0+   8.0  110.0 
#> [631]  46.0   14.0    1.0+ 145.0    3.0   24.0   50.0  126.0  180.0+   1.0 
#> [641] 178.0+  89.0   75.0    3.0+   1.0   33.0  180.0+ 168.0   52.0    7.0 
#> [651]   4.0    0.5  180.0+ 180.0+   4.0  180.0+  50.0  179.0+ 180.0+  16.0 
#> [661]  67.0   12.0    8.0   26.0    7.0+   0.5    2.0  180.0+   3.0   15.0+
#> 
#> $weights
#> NULL
#> 
#> $offset
#> NULL
#> 

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)
#> Warning: Multiple lambdas have been fit. Lambda will be set to 0.01 (see parameter 's').

# Score the predictions
predictions$score()
#> surv.cindex 
#>   0.8603997